Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-15 to 2019-02-22

Index       A        B        C        D
2018-11-15  -0.9852  -0.5294  -0.8167  0.1319
2018-11-16  -0.5694  -2.2496  -0.4365  0.0857
2018-11-17  0.0655   -0.8145  0.9066   0.5906
2018-11-18  0.0438   0.3349   0.1311   1.1015
2018-11-19  -1.1114  -0.0432  1.8993   0.5268
2018-11-20  -1.4863  0.8457   1.4789   -0.9561
2018-11-21  0.5784   -1.1013  0.3298   0.5362
⋮
2019-02-15  0.4545   0.5411   -1.0183  -0.3747
2019-02-16  0.2176   -0.7913  -0.0007  1.3458
2019-02-17  2.3082   0.0873   -0.8069  0.8815
2019-02-18  -1.3125  -0.3107  1.4388   -0.6217
2019-02-19  0.163    1.115    -1.0588  0.1472
2019-02-20  -0.4964  -1.8585  1.4133   0.3239
2019-02-21  -0.4473  1.3294   -0.0629  -0.7671
2019-02-22  -1.3048  1.1528   -0.4714  1.4845

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A        B        C        D
2018-11-30  -0.1694  -0.1965  0.2337   0.2985
2018-12-31  0.3709   -0.0935  -0.3007  -0.2434
2019-01-31  -0.0359  -0.0145  -0.0464  -0.0866

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
0.056584676890326605

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-15 to 2019-02-22

Index       A        B        C        D
2018-11-15  1.9888   0.7232   1.0204   -1.2581
2018-11-16  0.2901   -1.0253  -0.256   0.83
2018-11-17  -1.9609  0.353    0.1332   -2.2396
2018-11-18  -0.0625  0.6336   0.9737   0.9738
2018-11-19  0.7986   1.9573   -0.8355  -0.5254
2018-11-20  1.2609   -0.0154  -0.0755  -1.0475
2018-11-21  -1.0184  0.071    0.7955   -1.7798
⋮
2019-02-15  1.2463   1.6249   1.2064   0.9391
2019-02-16  -0.6299  -0.2194  -0.1274  -0.0425
2019-02-17  0.821    -0.3363  -0.9229  -0.6012
2019-02-18  0.7504   0.2714   -0.1805  -0.6694
2019-02-19  0.4193   -0.7244  0.3837   -0.9291
2019-02-20  -0.5081  -0.4752  -0.0197  -1.3213
2019-02-21  -0.7987  0.5301   0.5801   -0.1885
2019-02-22  0.4096   -0.7372  0.4937   -0.756

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-15 to 2019-02-22

Index       Sum
2018-11-15  2.4743
2018-11-16  -0.1611
2018-11-17  -3.7143
2018-11-18  2.5186
2018-11-19  1.3951
2018-11-20  0.1225
2018-11-21  -1.9317
⋮
2019-02-15  5.0167
2019-02-16  -1.0191
2019-02-17  -1.0394
2019-02-18  0.1719
2019-02-19  -0.8506
2019-02-20  -2.3243
2019-02-21  0.1231
2019-02-22  -0.5899

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 13.776  -4.13424  18.9393  -9.01409