PowerSystems

PowerSystems

API documentation

PowerSystems

Index

Exported

Internal

Index

PowerSystems.conductorless

PowerSystems.EconRenewable

PowerSystems.EconThermal

PowerSystems.MultiConductorMatrix

PowerSystems.MultiConductorValue

PowerSystems.MultiConductorVector

PowerSystems.ProportionalReserve

PowerSystems.StaticReserve

PowerSystems.TechThermal

PowerSystems.ThermalDispatch

PowerSystems.ThermalGen

PowerSystems.ThermalGenSeason

PowerSystems.Transformer2W

PowerSystems.VSCDCLine

Base.isapprox

Base.setindex!

Base.setindex!

InfrastructureModels._value2string

PowerSystems._check_conductors

PowerSystems._check_cost_functions

PowerSystems._check_reference_buses

PowerSystems._dfs

PowerSystems._make_mixed_units

PowerSystems._make_multiconductor

PowerSystems._make_per_unit

PowerSystems._propagate_topology_status

PowerSystems._rescale_cost_model

PowerSystems._select_largest_component

PowerSystems._simplify_pwl_cost

PowerSystems.add_dcline_costs

PowerSystems.add_section_data!

PowerSystems.apply_func

PowerSystems.bus_gen_lookup

PowerSystems.bus_load_lookup

PowerSystems.bus_shunt_lookup

PowerSystems.calc_branch_t

PowerSystems.calc_branch_y

PowerSystems.calc_theta_delta_bounds

PowerSystems.check_branch_directions

PowerSystems.check_branch_loops

PowerSystems.check_bus_types

PowerSystems.check_component_refrence_bus

PowerSystems.check_conductors

PowerSystems.check_connectivity

PowerSystems.check_cost_functions

PowerSystems.check_current_limits

PowerSystems.check_dcline_limits

PowerSystems.check_keys

PowerSystems.check_network_data

PowerSystems.check_reference_buses

PowerSystems.check_thermal_limits

PowerSystems.check_transformer_parameters

PowerSystems.check_voltage_angle_differences

PowerSystems.check_voltage_setpoints

PowerSystems.connected_components

PowerSystems.create_starbus_from_transformer

PowerSystems.export_cost_data

PowerSystems.export_extra_data

PowerSystems.export_matpower

PowerSystems.export_matpower

PowerSystems.find_max_bus_id

PowerSystems.get_bus_value

PowerSystems.get_default

PowerSystems.get_line_elements

PowerSystems.get_pti_dtypes

PowerSystems.get_pti_sections

PowerSystems.import_remaining!

PowerSystems.init_bus!

PowerSystems.make_mixed_units

PowerSystems.make_multiconductor

PowerSystems.make_per_unit

PowerSystems.matpower_to_powermodels

PowerSystems.merge_bus_name_data

PowerSystems.merge_generator_cost_data

PowerSystems.merge_generic_data

PowerSystems.mp2pm_branch

PowerSystems.mp2pm_dcline

PowerSystems.parse_file

PowerSystems.parse_json

PowerSystems.parse_json

PowerSystems.parse_line_element!

PowerSystems.parse_matpower

PowerSystems.parse_matpower_file

PowerSystems.parse_matpower_file

PowerSystems.parse_matpower_string

PowerSystems.parse_psse

PowerSystems.parse_psse

PowerSystems.parse_psse

PowerSystems.parse_pti

PowerSystems.parse_pti

PowerSystems.parse_pti_data

PowerSystems.print_summary

PowerSystems.propagate_topology_status

PowerSystems.psse2pm_branch!

PowerSystems.psse2pm_bus!

PowerSystems.psse2pm_dcline!

PowerSystems.psse2pm_generator!

PowerSystems.psse2pm_load!

PowerSystems.psse2pm_shunt!

PowerSystems.psse2pm_transformer!

PowerSystems.row_to_dict

PowerSystems.row_to_typed_dict

PowerSystems.select_largest_component

PowerSystems.split_loads_shunts

PowerSystems.standardize_cost_terms

PowerSystems.summary

PowerSystems.summary

PowerSystems.update_data

Exported

PowerSystems.EconRenewable — Type.

" Data Structure for the economical parameters of renewable generation technologies. The data structure can be called calling all the fields directly or using named fields. All the limits are defined by NamedTuples and some fields can take nothing

## Examples

PowerSystems.EconThermal — Type.

" Data Structure for the economical parameters of thermal generation technologies. The data structure can be called calling all the fields directly or using named fields. All the limits are defined by NamedTuples and some fields can take nothing

## Examples

PowerSystems.ProportionalReserve — Type.

ProportionalReserve(name::String, contributingdevices::PowerSystemDevice, timeframe::Float64, requirement::Dict{Any,Dict{Int,TimeSeries.TimeArray}})

Data Structure for a proportional reserve product for system simulations. The data structure can be called calling all the fields directly or using named fields. name - description contributingdevices - devices from which the product can be procured timeframe - the relative saturation timeframe requirement - the required quantity of the product

Examples

```jldoctest

PowerSystems.StaticReserve — Type.

StaticReserve(name::String, contributingdevices::PowerSystemDevice, timeframe::Float64, requirement::Float64})

Data Structure for the procurement products for system simulations. The data structure can be called calling all the fields directly or using named fields. name - description contributingdevices - devices from which the product can be procured timeframe - the relative saturation timeframe requirement - the required quantity of the product

Examples

```jldoctest

PowerSystems.TechThermal — Type.

TechThermal(activepower::Float64,
        activepowerlimits::NamedTuple{(:min, :max),Tuple{Float64,Float64}},
        reactivepower::Union{Float64,Nothing},
        reactivepowerlimits::Union{(min::Float64,max::Float64),Nothing},
        ramplimits::Union{NamedTuple{(:up, :down),Tuple{Float64,Float64}},Nothing},
        timelimits::Union{NamedTuple{(:min, :max),Tuple{Float64,Float64}},Nothing})

Data Structure for the economical parameters of thermal generation technologies. The data structure can be called calling all the fields directly or using named fields. Two examples are provided one with minimal data definition and a more comprenhensive one

# Examples

```jldoctest

julia> Tech = TechThermal(activepower = 100.0, activepowerlimits = (min = 50.0, max = 200.0))
WARNING: Limits defined as nothing
Tech Gen:
    Real Power: 100.0
    Real Power Limits: (min = 50.0, max = 200.0)
    Reactive Power: nothing
    Reactive Power Limits: nothing
    Ramp Limits: nothing
    Time Limits: nothing

PowerSystems.ThermalDispatch — Type.

" Data Structure for thermal generation technologies. The data structure contains all the information for technical and economical modeling. The data fields can be filled using named fields or directly.

Examples

PowerSystems.ThermalGen — Type.

Abstract struct for thermal generation technologies

PowerSystems.ThermalGenSeason — Type.

" Data Structure for thermal generation technologies subjecto to seasonality constraints. The data structure contains all the information for technical and economical modeling and an extra field for a time series. The data fields can be filled using named fields or directly.

Examples

PowerSystems.Transformer2W — Type.

The 2-W transformer model uses an equivalent circuit assuming the impedance is on the High Voltage Side of the transformer. The model allocates the iron losses and magnetezing suceptance to the primary side

PowerSystems.VSCDCLine — Type.

As implemented in Milano's Book Page 397

PowerSystems.parse_file — Method.

parse_file(file; import_all)

Parses a Matpower .m file or PTI (PSS(R)E-v33) .raw file into a PowerModels data structure. All fields from PTI files will be imported if import_all is true (Default: false).

Internal

PowerSystems.conductorless — Constant.

feild names that should not be multi-conductor values

PowerSystems.MultiConductorMatrix — Type.

PowerSystems.MultiConductorValue — Type.

a data structure for working with multiconductor datasets

PowerSystems.MultiConductorVector — Type.

a data structure for working with multiconductor datasets

Base.isapprox — Method.

Base.setindex! — Method.

Base.setindex! — Method.

InfrastructureModels._value2string — Method.

converts a MultiConductorValue value to a string in summary

PowerSystems._check_conductors — Method.

PowerSystems._check_cost_functions — Method.

PowerSystems._check_reference_buses — Method.

PowerSystems._dfs — Method.

performs DFS on a graph

PowerSystems._make_mixed_units — Method.

PowerSystems._make_multiconductor — Method.

PowerSystems._make_per_unit — Method.

PowerSystems._propagate_topology_status — Method.

PowerSystems._rescale_cost_model — Method.

PowerSystems._select_largest_component — Method.

PowerSystems._simplify_pwl_cost — Function.

checks the slope of each segment in a pwl function, simplifies the function if the slope changes is below a tolerance

PowerSystems.add_dcline_costs — Method.

adds dcline costs, if gen costs exist

PowerSystems.add_section_data! — Method.

add_section_data!(pti_data, section_data, section)

Adds section_data::Dict, which contains all parsed elements of a PTI file section given by section, into the parent pti_data::Dict

PowerSystems.apply_func — Method.

PowerSystems.bus_gen_lookup — Method.

builds a lookup list of what generators are connected to a given bus

PowerSystems.bus_load_lookup — Method.

builds a lookup list of what loads are connected to a given bus

PowerSystems.bus_shunt_lookup — Method.

builds a lookup list of what shunts are connected to a given bus

PowerSystems.calc_branch_t — Method.

PowerSystems.calc_branch_y — Method.

PowerSystems.calc_theta_delta_bounds — Method.

PowerSystems.check_branch_directions — Method.

checks that all parallel branches have the same orientation

PowerSystems.check_branch_loops — Method.

checks that all branches connect two distinct buses

PowerSystems.check_bus_types — Method.

checks bus types are consistent with generator connections, if not, fixes them

PowerSystems.check_component_refrence_bus — Method.

checks that a connected component has a reference bus, if not, adds one

PowerSystems.check_conductors — Method.

PowerSystems.check_connectivity — Method.

checks that all buses are unique and other components link to valid buses

PowerSystems.check_cost_functions — Method.

throws warnings if cost functions are malformed

PowerSystems.check_current_limits — Method.

checks that each branch has a reasonable current rating-a, if not computes one

PowerSystems.check_dcline_limits — Method.

checks that parameters for dc lines are reasonable

PowerSystems.check_keys — Method.

PowerSystems.check_network_data — Method.

PowerSystems.check_reference_buses — Method.

checks that each connected components has a reference bus, if not, adds one

PowerSystems.check_thermal_limits — Method.

checks that each branch has a reasonable thermal rating-a, if not computes one

PowerSystems.check_transformer_parameters — Method.

checks that each branch has a reasonable transformer parameters

this is important because setting tap == 0.0 leads to NaN computations, which are hard to @debug

PowerSystems.check_voltage_angle_differences — Function.

checks that voltage angle differences are within 90 deg., if not tightens

PowerSystems.check_voltage_setpoints — Method.

throws warnings if generator and dc line voltage setpoints are not consistent with the bus voltage setpoint

PowerSystems.connected_components — Method.

computes the connected components of the network graph returns a set of sets of bus ids, each set is a connected component

PowerSystems.create_starbus_from_transformer — Method.

create_starbus(pm_data, transformer)

Creates a starbus from a given three-winding transformer. "sourceid" is given by `["busi", "name", "I", "J", "K", "CKT"]` where "bus_i" and "name" are the modified names for the starbus, and "I", "J", "K" and "CKT" come from the originating transformer, in the PSS(R)E transformer specification.

PowerSystems.export_cost_data — Method.

Export cost data

PowerSystems.export_extra_data — Function.

Export fields of a component type

PowerSystems.export_matpower — Method.

Export power network data in the matpower format

PowerSystems.export_matpower — Method.

Export power network data in the matpower format

PowerSystems.find_max_bus_id — Method.

find_max_bus_id(pm_data)

Returns the maximum bus id in pm_data

PowerSystems.get_bus_value — Method.

get_bus_value(bus_i, field, pm_data)

Returns the value of field of bus_i from the PowerModels data. Requires "bus" Dict to already be populated.

PowerSystems.get_default — Function.

Get a default value for dict entry

PowerSystems.get_line_elements — Method.

get_line_elements(line)

Uses regular expressions to extract all separate data elements from a line of a PTI file and populate them into an Array{String}. Comments, typically indicated at the end of a line with a '/' character, are also extracted separately, and Array{Array{String}, String} is returned.

PowerSystems.get_pti_dtypes — Method.

get_pti_dtypes(field_name)

Returns OrderedDict of data types for PTI file section given by field_name, as enumerated by PSS/E Program Operation Manual

PowerSystems.get_pti_sections — Method.

get_pti_sections()

Returns Array of the names of the sections, in the order that they appear in a PTI file, v33+

PowerSystems.import_remaining! — Method.

Imports remaining keys from data_in into data_out, excluding keys in exclude

PowerSystems.init_bus! — Method.

init_bus!(bus, id)

Initializes a bus of id id with default values given in the PSS(R)E specification.

PowerSystems.make_mixed_units — Method.

Transforms network data into mixed-units (inverse of per-unit)

PowerSystems.make_multiconductor — Method.

Transforms single-conductor network data into multi-conductor data

PowerSystems.make_per_unit — Method.

Transforms network data into per-unit

PowerSystems.matpower_to_powermodels — Method.

Converts a Matpower dict into a PowerModels dict

PowerSystems.merge_bus_name_data — Method.

merges bus name data into buses, if names exist

PowerSystems.merge_generator_cost_data — Method.

merges generator cost functions into generator data, if costs exist

PowerSystems.merge_generic_data — Method.

merges Matpower tables based on the table extension syntax

PowerSystems.mp2pm_branch — Method.

sets all branch transformer taps to 1.0, to simplify branch models

PowerSystems.mp2pm_dcline — Method.

adds pmin and pmax values at to and from buses

PowerSystems.parse_json — Method.

PowerSystems.parse_json — Method.

PowerSystems.parse_line_element! — Method.

parse_line_element!(data, elements, section)

Parses a single "line" of data elements from a PTI file, as given by elements which is an array of the line, typically split at ,. Elements are parsed into data types given by section and saved into data::Dict

PowerSystems.parse_matpower — Method.

Parses the matpwer data from either a filename or an IO object

PowerSystems.parse_matpower_file — Method.

PowerSystems.parse_matpower_file — Method.

PowerSystems.parse_matpower_string — Method.

PowerSystems.parse_psse — Method.

parse_psse(pti_data)

Converts PSS(R)E-style data parsed from a PTI raw file, passed by pti_data into a format suitable for use internally in PowerModels. Imports all remaining data from the PTI file if import_all is true (Default: false).

PowerSystems.parse_psse — Method.

Parses directly from iostream

PowerSystems.parse_psse — Method.

Parses directly from file

PowerSystems.parse_pti — Method.

parse_pti(io::IO)

Reads PTI data in io::IO, returning a Dict of the data parsed into the proper types.

PowerSystems.parse_pti — Method.

parse_pti(filename::String)

Open PTI raw file given by filename, returning a Dict of the data parsed into the proper types.

PowerSystems.parse_pti_data — Method.

parse_pti_data(data_string, sections)

Parse a PTI raw file into a Dict, given the data_string of the file and a list of the sections in the PTI file (typically given by default by get_pti_sections().

PowerSystems.print_summary — Method.

prints the text summary for a data file or dictionary to STDOUT

PowerSystems.propagate_topology_status — Method.

finds active network buses and branches that are not necessary for the computation and sets their status to off.

Works on a PowerModels data dict, so that a it can be used without a GenericPowerModel object

Warning: this implementation has quadratic complexity, in the worst case

PowerSystems.psse2pm_branch! — Method.

psse2pm_branch!(pm_data, pti_data)

Parses PSS(R)E-style Branch data into a PowerModels-style Dict. "source_id" is given by ["I", "J", "CKT"] in PSS(R)E Branch specification.

PowerSystems.psse2pm_bus! — Method.

psse2pm_bus!(pm_data, pti_data)

Parses PSS(R)E-style Bus data into a PowerModels-style Dict. "source_id" is given by ["I", "NAME"] in PSS(R)E Bus specification.

PowerSystems.psse2pm_dcline! — Method.

psse2pm_dcline!(pm_data, pti_data)

Parses PSS(R)E-style Two-Terminal and VSC DC Lines data into a PowerModels compatible Dict structure by first converting them to a simple DC Line Model. For Two-Terminal DC lines, "sourceid" is given by ["IPR", "IPI", "NAME"] in the PSS(R)E Two-Terminal DC specification. For Voltage Source Converters, "sourceid" is given by ["IBUS1", "IBUS2", "NAME"], where "IBUS1" is "IBUS" of the first converter bus, and "IBUS2" is the "IBUS" of the second converter bus, in the PSS(R)E Voltage Source Converter specification.

PowerSystems.psse2pm_generator! — Method.

psse2pm_generator!(pm_data, pti_data)

Parses PSS(R)E-style Generator data in a PowerModels-style Dict. "source_id" is given by ["I", "ID"] in PSS(R)E Generator specification.

PowerSystems.psse2pm_load! — Method.

psse2pm_load!(pm_data, pti_data)

Parses PSS(R)E-style Load data into a PowerModels-style Dict. "source_id" is given by ["I", "ID"] in the PSS(R)E Load specification.

PowerSystems.psse2pm_shunt! — Method.

psse2pm_shunt!(pm_data, pti_data)

Parses PSS(R)E-style Fixed and Switched Shunt data into a PowerModels-style Dict. "source_id" is given by ["I", "ID"] for Fixed Shunts, and ["I", "SWREM"] for Switched Shunts, as given by the PSS(R)E Fixed and Switched Shunts specifications.

PowerSystems.psse2pm_transformer! — Method.

psse2pm_transformer!(pm_data, pti_data)

Parses PSS(R)E-style Transformer data into a PowerModels-style Dict. "source_id" is given by ["I", "J", "K", "CKT", "winding"], where "winding" is 0 if transformer is two-winding, and 1, 2, or 3 for three-winding, and the remaining keys are defined in the PSS(R)E Transformer specification.

PowerSystems.row_to_dict — Method.

takes a row from a matrix and assigns the values names

PowerSystems.row_to_typed_dict — Method.

takes a row from a matrix and assigns the values names and types

PowerSystems.select_largest_component — Method.

determines the largest connected component of the network and turns everything else off

PowerSystems.split_loads_shunts — Method.

split_loads_shunts(data)

Seperates Loads and Shunts in data under separate "load" and "shunt" keys in the PowerModels data format. Includes references to originating bus via "loadbus" and "shuntbus" keys, respectively.

PowerSystems.standardize_cost_terms — Method.

ensures all polynomial costs functions have at least three terms

PowerSystems.summary — Method.

prints the text summary for a data dictionary to IO

PowerSystems.summary — Method.

prints the text summary for a data file to IO

PowerSystems.update_data — Method.

recursively applies new_data to data, overwriting information