Docstrings

NLSolversBase.AbstractConstraints

NLSolversBase.AbstractObjective

NLSolversBase.BoxConstraints

NLSolversBase.ConstraintBounds

NLSolversBase.InplaceObjective

NLSolversBase.NLSolversBase

NLSolversBase.NonDifferentiable

NLSolversBase.NotInplaceObjective

NLSolversBase.OnceDifferentiable

NLSolversBase.OnceDifferentiableConstraints

NLSolversBase.TwiceDifferentiable

NLSolversBase.TwiceDifferentiableConstraints

NLSolversBase.TwiceDifferentiableHV

NLSolversBase._cb

NLSolversBase._clear_df!

NLSolversBase._clear_f!

NLSolversBase._clear_h!

NLSolversBase._clear_hv!

NLSolversBase._symm

NLSolversBase._symmetrize

NLSolversBase.alloc_DF

NLSolversBase.alloc_H

NLSolversBase.checked_chunk

NLSolversBase.clear!

NLSolversBase.df

NLSolversBase.df!_from_df

NLSolversBase.diffeqdiff_fdtype

NLSolversBase.eval

NLSolversBase.f!_from_f — Function.

f!_from_f(f, F::Abstractarray)

Return an inplace version of f

NLSolversBase.f_calls

NLSolversBase.fdf

NLSolversBase.fdf!_from_fdf

NLSolversBase.g_calls

NLSolversBase.gradient — Function.

Evaluates the gradient value at x

This does not update obj.DF or obj.x_df.

Get the most recently evaluated gradient of obj.

Get the ith element of the most recently evaluated gradient of obj.

NLSolversBase.gradient! — Function.

Evaluates the gradient value at x.

Stores the value in obj.DF.

NLSolversBase.gradient!! — Function.

Force (re-)evaluation of the gradient value at x.

Stores the value in obj.DF.

NLSolversBase.h!_from_h

NLSolversBase.h_calls

NLSolversBase.has_not_dep_symbol_in_ad

NLSolversBase.hessian — Function.

Get the most recently evaluated Hessian of obj

NLSolversBase.hessian!

NLSolversBase.hessian!!

NLSolversBase.hv_calls

NLSolversBase.hv_product

NLSolversBase.hv_product!

NLSolversBase.hv_product!!

NLSolversBase.include

NLSolversBase.ineqstr

NLSolversBase.is_finitediff

NLSolversBase.is_forwarddiff

NLSolversBase.jacobian — Function.

Get the most recently evaluated Jacobian of obj.

NLSolversBase.jacobian!

NLSolversBase.jacobian!!

NLSolversBase.make_df

NLSolversBase.make_f

NLSolversBase.make_fdf

NLSolversBase.nconstraints — Function.

nconstraints(bounds) -> nc

The number of linear/nonlinear constraint functions supplied by the user. This does not include bounds-constraints on variables.

See also: nconstraints_x.

NLSolversBase.nconstraints_x — Function.

nconstraints_x(bounds) -> nx

The number of "meaningful" constraints (not ±Inf) on the x coordinates.

See also: nconstraints.

NLSolversBase.only_fg

NLSolversBase.only_fg!

NLSolversBase.only_fgh!

NLSolversBase.only_fj

NLSolversBase.only_fj!

NLSolversBase.only_g_and_fg

NLSolversBase.only_j_and_fj

NLSolversBase.parse_constraints — Function.

parse_constraints(T, l, u) -> eq, val, ineq, σ, b

From user-supplied constraints of the form

l_i ≤  v_i  ≤ u_i

(which include both inequality and equality constraints, the latter when l_i == u_i), convert into the following representation:

- `eq`, a vector of the indices for which `l[eq] == u[eq]`
- `val = l[eq] = u[eq]`
- `ineq`, `σ`, and `b` such that the inequality constraints can be written as
         σ[k]*(v[ineq[k]] - b[k]) ≥ 0
   where `σ[k] = ±1`.

Note that since the same v_i might have both lower and upper bounds, ineq might have the same index twice (once with σ=-1 and once with σ=1).

Supplying ±Inf for elements of l and/or u implies that v_i is unbounded in the corresponding direction. In such cases there is no corresponding entry in ineq/σ/b.

T is the element-type of the non-Int outputs

NLSolversBase.showeq

NLSolversBase.showineq

NLSolversBase.symmetrize

NLSolversBase.value — Function.

Evaluates the objective value at x.

Returns f(x), but does not store the value in obj.F

Get the most recently evaluated objective value of obj.

NLSolversBase.value! — Function.

Evaluates the objective value at x.

Returns f(x) and stores the value in obj.F

NLSolversBase.value!! — Function.

Force (re-)evaluation of the objective value at x.

Returns f(x) and stores the value in obj.F

NLSolversBase.value_gradient!

NLSolversBase.value_gradient!!

NLSolversBase.value_jacobian!

NLSolversBase.value_jacobian!!

NLSolversBase.x_of_nans