Joins

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-16 to 2019-02-22

Index       A
2018-06-16  0.5984
2018-06-17  0.7559
2018-06-18  0.8766
2018-06-19  0.0362
2018-06-20  0.9471
2018-06-21  0.7443
2018-06-22  0.5693
⋮
2019-02-15  0.5813
2019-02-16  0.5243
2019-02-17  0.8744
2019-02-18  0.9848
2019-02-19  0.912
2019-02-20  0.0428
2019-02-21  0.3093
2019-02-22  0.3505

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-16 to 2018-08-22

Index       A
2017-12-16  0.3323
2017-12-17  0.0682
2017-12-18  0.4348
2017-12-19  0.0472
2017-12-20  0.8604
2017-12-21  0.5127
2017-12-22  0.5629
⋮
2018-08-15  0.2068
2018-08-16  0.1506
2018-08-17  0.4856
2018-08-18  0.2929
2018-08-19  0.6103
2018-08-20  0.0988
2018-08-21  0.6901
2018-08-22  0.0305

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-16 to 2019-02-22

Index       A       A
2017-12-16  NaN     0.3323
2017-12-17  NaN     0.0682
2017-12-18  NaN     0.4348
2017-12-19  NaN     0.0472
2017-12-20  NaN     0.8604
2017-12-21  NaN     0.5127
2017-12-22  NaN     0.5629
⋮
2019-02-15  0.5813  NaN
2019-02-16  0.5243  NaN
2019-02-17  0.8744  NaN
2019-02-18  0.9848  NaN
2019-02-19  0.912   NaN
2019-02-20  0.0428  NaN
2019-02-21  0.3093  NaN
2019-02-22  0.3505  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-16 to 2019-02-22

Index       A
2018-06-16  0.7549
2018-06-17  0.1949
2018-06-18  0.6834
2018-06-19  0.4055
2018-06-20  0.0957
2018-06-21  0.7726
2018-06-22  0.0089
⋮
2019-02-15  0.8318
2019-02-16  0.0968
2019-02-17  0.0484
2019-02-18  0.6965
2019-02-19  0.8349
2019-02-20  0.586
2019-02-21  0.1636
2019-02-22  0.0077

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-16 to 2018-08-22

Index       A
2017-12-16  0.6139
2017-12-17  0.7745
2017-12-18  0.1098
2017-12-19  0.4311
2017-12-20  0.6966
2017-12-21  0.8215
2017-12-22  0.2887
⋮
2018-08-15  0.8281
2018-08-16  0.3208
2018-08-17  0.1802
2018-08-18  0.5527
2018-08-19  0.8329
2018-08-20  0.2278
2018-08-21  0.095
2018-08-22  0.0589

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-16 to 2019-02-22

Index       A
2018-06-16  0.1142
2018-06-17  0.3491
2018-06-18  0.3944
2018-06-19  0.3481
2018-06-20  0.962
2018-06-21  0.4737
2018-06-22  0.2044
⋮
2019-02-15  0.4791
2019-02-16  0.8892
2019-02-17  0.8856
2019-02-18  0.691
2019-02-19  0.765
2019-02-20  0.7336
2019-02-21  0.8847
2019-02-22  0.7119

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-16 to 2018-08-22

Index       A
2017-12-16  0.174
2017-12-17  0.4138
2017-12-18  0.1359
2017-12-19  0.8297
2017-12-20  0.8748
2017-12-21  0.9102
2017-12-22  0.282
⋮
2018-08-15  0.5369
2018-08-16  0.903
2018-08-17  0.6691
2018-08-18  0.3095
2018-08-19  0.9036
2018-08-20  0.61
2018-08-21  0.9798
2018-08-22  0.2169

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-16 to 2019-02-22

Index       A       A
2018-06-16  0.1142  0.2714
2018-06-17  0.3491  0.013
2018-06-18  0.3944  0.7669
2018-06-19  0.3481  0.6984
2018-06-20  0.962   0.5665
2018-06-21  0.4737  0.1693
2018-06-22  0.2044  0.9542
⋮
2019-02-15  0.4791  NaN
2019-02-16  0.8892  NaN
2019-02-17  0.8856  NaN
2019-02-18  0.691   NaN
2019-02-19  0.765   NaN
2019-02-20  0.7336  NaN
2019-02-21  0.8847  NaN
2019-02-22  0.7119  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-16 to 2018-08-22

Index       A       A
2017-12-16  NaN     0.174
2017-12-17  NaN     0.4138
2017-12-18  NaN     0.1359
2017-12-19  NaN     0.8297
2017-12-20  NaN     0.8748
2017-12-21  NaN     0.9102
2017-12-22  NaN     0.282
⋮
2018-08-15  0.8874  0.0471
2018-08-16  0.4181  0.3979
2018-08-17  0.7305  0.2094
2018-08-18  0.6011  0.2248
2018-08-19  0.6175  0.1785
2018-08-20  0.581   0.5369
2018-08-21  0.0439  0.903
2018-08-22  0.6744  0.6691