Amplitude adjusted Fourier (AAFT)

Amplitude adjusted Fourier transform surrogates

Different variants of the AAFT and iterated AAFT algorithms.

Amplitude adjusted Fourier transform (AAFT)

TimeseriesSurrogates.aaft — Function.

aaft(ts::AbstractArray{T, 1} where T)

Generate a realization of an amplitude adjusted Fourier transform (AAFT) surrogate (Theiler et al., 1992).

Arguments

ts: the time series for which to generate the surrogate realization.

References

J. Theiler et al., Physica D 58 (1992) 77-94 (1992). https://www.sciencedirect.com/science/article/pii/016727899290102S

Example of AAFT

using TimeseriesSurrogates
ts = AR1() # create a realization of a random AR(1) process
surrogate = aaft(ts)

surrplot(ts, surrogate)

Iterated AAFT (AAFT)

TimeseriesSurrogates.iaaft — Function.

iaaft(ts::AbstractArray{T, 1} where T;
        n_maxiter = 200, tol = 1e-6, n_windows = 50)

Generate an iteratively adjusted amplitude adjusted Fourier transform (IAAFT) (Schreiber & Schmitz, 1996) surrogate realization of ts.

Arguments

ts: the time series for which to generate an AAFT surrogate realization.

n_maxiter: sets the maximum number of iterations to allow before ending the algorithm (if convergence is slow).

tol: the relative tolerance for deciding if convergence is achieved.

n_window: the number is windows used when binning the periodogram (used for determining convergence).

References

T. Schreiber; A. Schmitz (1996). "Improved Surrogate Data for Nonlinearity Tests". Phys. Rev. Lett. 77 (4): 635–638. doi:10.1103/PhysRevLett.77.635. PMID 10062864. https://journals.aps.org/prl/abstract/10.1103/PhysRevLett.77.635

The IAAFT surrogates add an iterative step to the AAFT algorithm improve convergence.

Example of IAAFT

using TimeseriesSurrogates
ts = AR1() # create a realization of a random AR(1) process
surrogate = iaaft(ts)

surrplot(ts, surrogate)