Grid estimator

Grid estimator

PerronFrobenius.TransferOperators.Estimators.transferoperator — Function.

transferoperator(points, binning_scheme::RectangularBinningScheme; kwargs...)

Discretize points using the provided binning_scheme and compute the transfer operator over the partition elements.

transferoperator(pts, ϵ::TriangulationBinning, simplex_intersection_type::ApproximateIntersection; 
    n::Int = 200, sample_randomly::Bool = false) -> TransferOperatorTriangulationApprox

Estimate the invariant measure over the state space defined by pts using a triangulation of the phase space as the partition, using approximate simplex intersections to compute transition probabilities between the states (simplices).

n is the number of points that each simplex is sampled with, and sample_randomly indicates whether points used be sampled randomly within each simpled (sample_randomly = true) or by a regular simplex splitting routine (sample_randomly = false, which is default).

Example: If pts = [rand(3) for i = 1:30], then run transferoperator(pts, TriangulationBinning(), ApproximateIntersection()).

transferoperator(pts, ϵ::TriangulationBinning, 
    simplex_intersection_type::ExactIntersection) -> TransferOperatorTriangulationExact

Estimate the invariant measure over the state space defined by pts using a triangulation of the phase space as the partition, using exact simplex intersections to compute transition probabilities between the states (simplices).

Example: If pts = [rand(3) for i = 1:30], then run transferoperator(pts, TriangulationBinning(), ExactIntersection()).

transferoperator(points, binning_scheme::TriangulationBinningScheme,
    simplex_intersection::SimplexIntersectionType; kwargs...)

Discretize points using the provided binning_scheme and compute the transfer operator over the partition elements using the type of simplex intersections indicated by simplex_intersection_type.