Compilation & Execution

Compilation & Execution

CUDAnative.@cuda — Macro.

@cuda [kwargs...] func(args...)

High-level interface for executing code on a GPU. The @cuda macro should prefix a call, with func a callable function or object that should return nothing. It will be compiled to a CUDA function upon first use, and to a certain extent arguments will be converted and managed automatically using cudaconvert. Finally, a call to CUDAdrv.cudacall is performed, scheduling a kernel launch on the current CUDA context.

Several keyword arguments are supported that influence kernel compilation and execution. For more information, refer to the documentation of respectively cufunction and CUDAnative.Kernel

The underlying operations (argument conversion, kernel compilation, kernel call) can be performed explicitly when more control is needed, e.g. to reflect on the resource usage of a kernel to determine the launch configuration:

args = ...
GC.@preserve args begin
    kernel_args = cudaconvert.(args)
    kernel_tt = Tuple{Core.Typeof.(kernel_args)...}
    kernel = cufunction(f, kernel_tt; compilation_kwargs)
    kernel(kernel_args...; launch_kwargs)
end

CUDAnative.cufunction — Function.

cufunction(f, tt=Tuple{}; kwargs...)

Low-level interface to compile a function invocation for the currently-active GPU, returning a callable kernel object. For a higher-level interface, use @cuda.

The following keyword arguments are supported:

minthreads: the required number of threads in a thread block.

maxthreads: the maximum number of threads in a thread block.

blockspersm: a minimum number of thread blocks to be scheduled on a single multiprocessor.

maxregs: the maximum number of registers to be allocated to a single thread (only supported on LLVM 4.0+)

The output of this function is automatically cached, i.e. you can simply call cufunction in a hot path without degrading performance. New code will be generated automatically, when when function changes, or when different types or keyword arguments are provided.

CUDAnative.Kernel — Type.

(::Kernel)(args...; kwargs...)

Low-level interface to call a compiled kernel, passing GPU-compatible arguments in args. For a higher-level interface, use @cuda.

The following keyword arguments are supported:

threads (defaults to 1)

blocks (defaults to 1)

shmem (defaults to 0)

stream (defaults to the default stream)

CUDAnative.compile — Function.

compile(dev::CuDevice, f, tt; kwargs...)

Compile a function f invoked with types tt for device dev, returning the compiled function module respectively of type CuFuction and CuModule.

For a list of supported keyword arguments, refer to the documentation of cufunction.

CUDAnative.cudaconvert — Function.

cudaconvert(x)

This function is called for every argument to be passed to a kernel, allowing it to be converted to a GPU-friendly format. By default, the function does nothing and returns the input object x as-is.

Do not add methods to this function, but instead extend the underlying Adapt.jl package and register methods for the the CUDAnative.Adaptor type.

CUDAnative.nearest_warpsize — Function.

nearest_warpsize(dev::CuDevice, threads::Integer)

Return the nearest number of threads that is a multiple of the warp size of a device.

This is a common requirement, eg. when using shuffle intrinsics.

Devices

CUDAnative.device! — Function.

device!(dev)

Sets dev as the current active device for the calling host thread. Devices can be specified by integer id, or as a CuDevice. This is intended to be a low-cost operation, only performing significant work when calling it for the first time for each device.

If your library or code needs to perform an action when the active device changes, add a callback of the signature (::CuDevice, ::CuContext) to the device!_listeners set.

device!(f, dev)

Sets the active device for the duration of f.