Intrinsics

Intrinsics

This section lists the package's public functionality that corresponds to special CUDA functions to be used in device code. It is loosely organized according to the C language extensions appendix from the CUDA C programming guide. For more information about certain intrinsics, refer to the aforementioned NVIDIA documentation.

Indexing and Dimensions

CUDAnative.gridDim — Function.

gridDim()::CuDim3

Returns the dimensions of the grid.

CUDAnative.blockIdx — Function.

blockIdx()::CuDim3

Returns the block index within the grid.

CUDAnative.blockDim — Function.

blockDim()::CuDim3

Returns the dimensions of the block.

CUDAnative.threadIdx — Function.

threadIdx()::CuDim3

Returns the thread index within the block.

CUDAnative.warpsize — Function.

warpsize()::UInt32

Returns the warp size (in threads).

Memory Types

Shared Memory

CUDAnative.@cuStaticSharedMem — Macro.

@cuStaticSharedMem(T::Type, dims) -> CuDeviceArray{T,AS.Shared}

Get an array of type T and dimensions dims (either an integer length or tuple shape) pointing to a statically-allocated piece of shared memory. The type should be statically inferable and the dimensions should be constant, or an error will be thrown and the generator function will be called dynamically.

CUDAnative.@cuDynamicSharedMem — Macro.

@cuDynamicSharedMem(T::Type, dims, offset::Integer=0) -> CuDeviceArray{T,AS.Shared}

Get an array of type T and dimensions dims (either an integer length or tuple shape) pointing to a dynamically-allocated piece of shared memory. The type should be statically inferable or an error will be thrown and the generator function will be called dynamically.

Note that the amount of dynamic shared memory needs to specified when launching the kernel.

Optionally, an offset parameter indicating how many bytes to add to the base shared memory pointer can be specified. This is useful when dealing with a heterogeneous buffer of dynamic shared memory; in the case of a homogeneous multi-part buffer it is preferred to use view.

Synchronization

CUDAnative.sync_threads — Function.

sync_threads()

Waits until all threads in the thread block have reached this point and all global and shared memory accesses made by these threads prior to sync_threads() are visible to all threads in the block.

CUDAnative.sync_warp — Function.

sync_warp(mask::Integer=0xffffffff)

Waits threads in the warp, selected by means of the bitmask mask, have reached this point and all global and shared memory accesses made by these threads prior to sync_warp() are visible to those threads in the warp. The default value for mask selects all threads in the warp.

Warp Vote

The warp vote functions allow the threads of a given warp to perform a reduction-and-broadcast operation. These functions take as input a boolean predicate from each thread in the warp and evaluate it. The results of that evaluation are combined (reduced) across the active threads of the warp in one different ways, broadcasting a single return value to each participating thread.

CUDAnative.vote_all — Function.

vote_all(predicate::Bool)

Evaluate predicate for all active threads of the warp and return non-zero if and only if predicate evaluates to non-zero for all of them.

CUDAnative.vote_any — Function.

vote_any(predicate::Bool)

Evaluate predicate for all active threads of the warp and return non-zero if and only if predicate evaluates to non-zero for any of them.

CUDAnative.vote_ballot — Function.

vote_ballot(predicate::Bool)

Evaluate predicate for all active threads of the warp and return an integer whose Nth bit is set if and only if predicate evaluates to non-zero for the Nth thread of the warp and the Nth thread is active.

Warp Shuffle

CUDAnative.shfl — Function.

shfl(val, lane::Integer, width::Integer=32)

Shuffle a value from a directly indexed lane lane.

CUDAnative.shfl_up — Function.

shfl_up(val, delta::Integer, width::Integer=32)

Shuffle a value from a lane with lower ID relative to caller.

CUDAnative.shfl_down — Function.

shfl_down(val, delta::Integer, width::Integer=32)

Shuffle a value from a lane with higher ID relative to caller.

CUDAnative.shfl_xor — Function.

shfl_xor(val, mask::Integer, width::Integer=32)

Shuffle a value from a lane based on bitwise XOR of own lane ID with mask.

If using CUDA 9.0, and PTX ISA 6.0 is supported, synchronizing versions of these intrinsics are available as well:

CUDAnative.shfl_sync — Function.

shfl_sync(val, lane::Integer, width::Integer=32, threadmask::UInt32=0xffffffff)

Shuffle a value from a directly indexed lane lane. The default value for threadmask performs the shuffle on all threads in the warp.

CUDAnative.shfl_up_sync — Function.

shfl_up_sync(val, delta::Integer, width::Integer=32, threadmask::UInt32=0xffffffff)

Shuffle a value from a lane with lower ID relative to caller. The default value for threadmask performs the shuffle on all threads in the warp.

CUDAnative.shfl_down_sync — Function.

shfl_down_sync(val, delta::Integer, width::Integer=32, threadmask::UInt32=0xffffffff)

Shuffle a value from a lane with higher ID relative to caller. The default value for threadmask performs the shuffle on all threads in the warp.

CUDAnative.shfl_xor_sync — Function.

shfl_xor_sync(val, mask::Integer, width::Integer=32, threadmask::UInt32=0xffffffff)

Shuffle a value from a lane based on bitwise XOR of own lane ID with mask. The default value for threadmask performs the shuffle on all threads in the warp.

Formatted Output

CUDAnative.@cuprintf — Macro.

Print a formatted string in device context on the host standard output:

@cuprintf("%Fmt", args...)

Note that this is not a fully C-compliant printf implementation; see the CUDA documentation for supported options and inputs.

Also beware that it is an untyped, and unforgiving printf implementation. Type widths need to match, eg. printing a 64-bit Julia integer requires the %ld formatting string.

Assertions

CUDAnative.@cuassert — Macro.

@assert cond [text]

Signal assertion failure to the CUDA driver if cond is false. Preferred syntax for writing assertions, mimicking Base.@assert. Message text is optionally displayed upon assertion failure.

Warning

A failed assertion will crash the GPU, so use sparingly as a debugging tool. Furthermore, the assertion might be disabled at various optimization levels, and thus should not cause any side-effects.