Internals

Internal APIs

Documentation for SpinMonteCarlo.jl's internals (not exported).

Driver

SpinMonteCarlo.accumulateObservables! — Function.

accumulateObservables!(model, obs::MCObservableSet, localobs::Dict)

Accumulates localobs into obs. For example, obs["Energy"] << localobs["Energy"].

SpinMonteCarlo.postproc — Function.

postproc(model::Model, param::Parameter, obs::MCObservableSet)

Post process of observables. For example, Specific heat will be calculated from energy, energy^2, and temperature.

SpinMonteCarlo.postproc — Method.

postproc(model::Union{Ising, Potts}, param::Parameter, obs::MCObservableSet)

Observables to be calculated

In the following, $m$ is total magnetization per site and $\epsilon$ is total energy per site.

"Binder Ratio"

\[R := \frac{\left \langle m^4 \right \rangle}{\left \langle m^2 \right\rangle^2}\]

"Susceptibility"

\[\chi := \frac{N}{T}\left(\left\langle m^2\right\rangle\right)\]

"Connected Susceptibility"

\[\frac{N}{T}\left(\left\langle m^2\right\rangle - \left\langle |m| \right\rangle^2\right)\]

"Specific Heat"

\[\frac{N}{T^2}\left(\left\langle \epsilon^2\right\rangle - \left\langle \epsilon \right\rangle^2\right)\]

SpinMonteCarlo.postproc — Method.

postproc(model::Union{Clock, XY}, param::Parameter, obs::MCObservableSet)

Observables to be calculated

In the following, $m$ is total magnetization per site and $\epsilon$ is total energy per site.

"Binder Ratio x"

\[\frac{\left \langle m_x^4 \right \rangle}{\left \langle m_x^2 \right\rangle^2}\]

"Binder Ratio y"

\[\frac{\left \langle m_y^4 \right \rangle}{\left \langle m_y^2 \right\rangle^2}\]

"Binder Ratio"

\[\frac{\left \langle |m|^4 \right \rangle}{\left \langle |m|^2 \right\rangle^2}\]

"Susceptibility x"

\[\frac{N}{T}\left(\left\langle m_x^2\right\rangle\right)\]

"Susceptibility y"

\[\frac{N}{T}\left(\left\langle m_y^2\right\rangle\right)\]

"Susceptibility y"

\[\frac{N}{T}\left(\left\langle |m|^2\right\rangle\right)\]

"Connected Susceptibility x"

\[\frac{N}{T}\left(\left\langle m_x^2\right\rangle - \left\langle |m_x| \right\rangle^2\right)\]

"Connected Susceptibility y"

\[\frac{N}{T}\left(\left\langle m_y^2\right\rangle - \left\langle |m_y| \right\rangle^2\right)\]

"Connected Susceptibility"

\[\frac{N}{T}\left(\left\langle |m|^2\right\rangle - \left\langle |m| \right\rangle^2\right)\]

"Specific Heat"

\[\frac{N}{T^2}\left(\left\langle \epsilon^2\right\rangle - \left\langle \epsilon \right\rangle^2\right)\]

SpinMonteCarlo.postproc — Method.

postproc(model::QuantumXXZ, param::Parameter, obs::MCObservableSet)

Observables to be calculated

In the following, $s$ is sign of weight, $m$ is total magnetization per site, and $\epsilon$ is total energy per site.

"Magnetization"

\[\left\langle m s\right\rangle\Big/\left\langle s \right\rangle\]

"|Magnetization|"

\[\left\langle |m| s\right\rangle\Big/\left\langle s \right\rangle\]

"Magnetization^2"

\[\left\langle m^2 s\right\rangle\Big/\left\langle s \right\rangle\]

"Magnetization^4"

\[\left\langle m^4 s\right\rangle\Big/\left\langle s \right\rangle\]

"Energy"

\[\left\langle \epsilon s\right\rangle\Big/\left\langle s \right\rangle\]

"Energy^2"

\[\left\langle \epsilon^2 s\right\rangle\Big/\left\langle s \right\rangle\]

"Binder Ratio"

\[\frac{\left \langle m^4 \right \rangle}{\left \langle m^2 \right\rangle^2}\]

"Susceptibility"

\[\frac{N}{T}\left(\left\langle m^2\right\rangle\right)\]

"Connected Susceptibility"

\[\frac{N}{T}\left(\left\langle m^2\right\rangle - \left\langle |m| \right\rangle^2\right)\]

"Specific Heat"

\[\frac{N}{T^2}\left(\left\langle \epsilon^2\right\rangle - \left\langle \epsilon \right\rangle^2\right)\]

Lattice

SpinMonteCarlo.Lattice — Type.

Lattice

Fields

dim :: Int

dimension of lattice

size :: Vector{Int}

length of lattice in each dimension

nsitetypes :: Int

the number of sitetypes

nbondtypes :: Int

the number of bondtypes

sites :: Vector{Vector{Int}}

list of site indecies for each sitetype

bonds :: Vector{Vector{Int}}

list of bond indecies for each sitetype

nsites :: Int

the number of sites

nbonds :: Int

the number of bonds

sitetypes :: Vector{Int}

sitetype of each site

bondtypes :: Vector{Int}

bondtype of each bond

transvector :: Matrix{Float64}

lattice vector represented in Cartesian system

@assert size(transvector) == (dim, dim)

site_coords :: Matrix{Float64}

coordinate of each site represented in lattice system

@assert size(site_coords) == (dim, nsites)

bond_dirs :: Matrix{Float64}

displacement of each bond represented in lattice system

@assert size(bond_dirs) == (dim, nbonds)

neighborsites :: Vector{Vector{Int}}

list of indecies of neighbor sites of each site

neighborbonds :: Vector{Vector{Int}}

list of indecies of neighbor bonds of each site

source :: Vector{Int}

the index of an end site of each bond index

target :: Vector{Int}

the index of the other end site of each bond index

Base.size — Method.

size(lat::Lattice, [dim::Integer])
size(model::Model, [dim::Integer])

Returns the size of lattice.

PDMats.dim — Method.

dim(lat::Lattice)
dim(model::Model)

Returns the dimension of lattice.

SpinMonteCarlo.bonddirection — Method.

bonddirection(lat::Lattice, bond::Integer)
bonddirection(model::Model, bond::Integer)

Returns the direction of the bond as vector in the Cartesian system

SpinMonteCarlo.bondtype — Method.

bondtype(lat::Lattice, bond::Integer)
bondtype(model::Model, bond::Integer)

Returns the type of bond

SpinMonteCarlo.lattice_bonddirection — Method.

lattice_bonddirection(lat::Lattice, bond::Integer)
lattice_bonddirection(model::Model, bond::Integer)

Returns the direction of the bond as vector in the lattice system

SpinMonteCarlo.lattice_sitecoordinate — Method.

lattice_sitecoordinate(lat::Lattice, site::Integer)
lattice_sitecoordinate(model::Model, site::Integer)

Returns the coordinate of the site in the lattice system

SpinMonteCarlo.neighbors — Method.

neighbors(lat::Lattice, site::Integer)
neighbors(model::Model, site::Integer)

Returns the neighbor sites and bonds of site.

SpinMonteCarlo.numbonds — Method.

numbonds(lat::Lattice, bondtype::Integer)
numbonds(model::Model, bondtype::Integer)

Returns the number of bondtype bonds.

SpinMonteCarlo.numbonds — Method.

numbonds(lat::Lattice)
numbonds(model::Model)

Returns the number of all bonds.

SpinMonteCarlo.numsites — Method.

numsites(lat::Lattice, sitetype::Integer)
numsites(model::Model, sitetype::Integer)

Returns the number of sitetype sites.

SpinMonteCarlo.numsites — Method.

numsites(lat::Lattice)
numsites(model::Model)

Returns the number of all sites.

SpinMonteCarlo.sitecoordinate — Method.

sitecoordinate(lat::Lattice, site::Integer)
sitecoordinate(model::Model, site::Integer)

Returns the coordinate of the site in the Cartesian system

SpinMonteCarlo.sitetype — Method.

sitetype(lat::Lattice, site::Integer)
sitetype(model::Model, site::Integer)

Returns the type of site

SpinMonteCarlo.source — Method.

source(lat::Lattice, bond::Integer)
source(model::Model, bond::Integer)

Returns the source site of bond.

SpinMonteCarlo.target — Method.

target(lat::Lattice, bond::Integer)
target(model::Model, bond::Integer)

Returns the target site of bond.

SpinMonteCarlo.bonds — Method.

bonds(lat::Lattice, bondtype::Integer)
bonds(model::Model, bondtype::Integer)

Returns bonds with bondtype

SpinMonteCarlo.neighborbonds — Method.

neighborbonds(lat::Lattice, site::Integer)
neighborbonds(model::Model, site::Integer)

Returns the neighbor bonds of site.

SpinMonteCarlo.neighborsites — Method.

neighborsites(lat::Lattice, site::Integer)
neighborsites(model::Model, site::Integer)

Returns the neighbor sites of site.

SpinMonteCarlo.numbondtypes — Method.

numbondtypes(lat::Lattice)
numbondtypes(model::Model)

Returns the number of bondtypes.

SpinMonteCarlo.numsitetypes — Method.

numsitetypes(lat::Lattice)
numsitetypes(model::Model)

Returns the number of sitetypes.

SpinMonteCarlo.sites — Method.

sites(lat::Lattice, sitetype::Integer)
sites(model::Model, sitetype::Integer)

Returns sites with sitetype

Model

SpinMonteCarlo.LoopElementType — Type.

Enumtype including LET_*

source

SpinMonteCarlo.LET_Cut — Constant.

Loop element depicted as

|
o
|

or matrix

1 1 |+>
1 1 |->

SpinMonteCarlo.LET_FMLink — Constant.

Loop element depicted as

|  |
|--|
|  |

or matrix

1 0 0 0 |++>
0 0 0 0 |+->
0 0 0 0 |-+>
0 0 0 1 |-->

SpinMonteCarlo.LET_AFLink — Constant.

Loop element depicted as

|  |
|~~|
|  |

or matrix

0 0 0 0 |++>
0 1 0 0 |+->
0 0 1 0 |-+>
0 0 0 0 |-->

SpinMonteCarlo.LET_Vertex — Constant.

Loop element depicted as

|  |
~~~~
~~~~
|  |

or matrix

0 0 0 0 |++>
0 1 1 0 |+->
0 1 1 0 |-+>
0 0 0 0 |-->

SpinMonteCarlo.LET_Cross — Constant.

Loop element depicted as

|  |
 \/
 /\
|  |

or matrix

1 0 0 0 |++>
0 0 1 0 |+->
0 1 0 0 |-+>
0 0 0 1 |-->

SpinMonteCarlo.LocalLoopOperator — Type.

(Imaginary-temporary and spatial) local operator as a perturbation with assigned loop element.

Fields

let_type : assigned loop element

isdiagonal : operator is diagonal or not

in other words, two states connecting this perturbation are equivalent to each other or not.

time : imaginary time ($\tau/\beta \in [0,1)$) which this perturbation acts on.

space : index of subspin or subbond which this perturbation acts on.

bottom_id :: index of node of union find assigned to a loop

top_id :: index of node of union find assigned to the other loop

Utility

SpinMonteCarlo.default_estimator — Function.

default_estimator(model, updatemethod!)

Determines estimator to be used when param["Estimator"] is not set.

SpinMonteCarlo.@gen_convert_parameter — Macro.

@gen_convert_parameter(model_typename, (keyname, size_fn, default)...)

Generates convert_parameter(model::model_typename, param::Parameter).

Example

@gen_convert_parameter(A, ("A", numbondtypes, 1.0), ("B", 1, 1))

generates a function equivalent to the following:

doc"""
    convert_parameter(model::A, param::Parameter)

# Keynames

- "A": a vector with `numbondtypes(model)` elements (default: 1)
- "B": a scalar (default: 1.0)

"""
function convert_parameter(model::A, param::Parameter)
    ## if `size_fn` is a `Function`,
    ## result is a vector whose size is `size_fn(model)`.
    ## `param["A"]` can take a scalar or a vector.
    a = get(param, "A", 1.0)
    as = zeros(Float64, numbondtypes(model))
    as .= a

    ## otherwise,
    ## result is a scalar.
    b = get(param, "B", 1) :: Int

    return as, b
end

SpinMonteCarlo.UnionFind — Type.

Union-find algorithm.

SpinMonteCarlo.addnode! — Function.

addnode!(u::UnionFind)

Adds a new node into u and returns the number of nodes including the added node.

SpinMonteCarlo.unify! — Function.

unify!(u, n1, n2)

Connects n1 and n2 nodes and returns the root.

SpinMonteCarlo.clusterize! — Function.

clusterize!(u::UnionFind)

Assigns cluster ID to each node and returns the number of clusters.

SpinMonteCarlo.clusterid — Function.

clusterid(u::UnionFind, i::Integer)

Returns the index of the cluster where i node belongs.