Example systems

Example processes

TimeseriesSurrogates.NLNS — Function.

NLNS(n_steps, x₀)

A nonlinear, nonstationary process.

Equations

The systems is given by the following stochastic model:

\[x(t+1) = x(t) + a(t) a(t-1)\]

with $a(t) extasciitilde N(0,1)$ being stationary. The nonlinearity lies in the $a(t) a(t-1)$ term. This is, in essence, a nonlinear random walk. x₀ sets the initial condition.

This example is from Lucio et al. (2012).

References

Lucio et al., Phys. Rev. E 85, 056202 (2012), after J. Timmer, Phys. Rev. E 58, 5153 (1998). https://journals.aps.org/pre/abstract/10.1103/PhysRevE.85.056202

TimeseriesSurrogates.SNLST — Function.

SNLST(n_steps, x₀, k)

Dynamically linear process transformed by a strongly nonlinear static transformation (SNLST), from Lucio et al., 2012.

Equations

The system is by the following map:

\[x(t) = k x(t-1) + a(t)\]

with the transformation $s(t) = x(t)^3$.

References

Lucio et al., Phys. Rev. E 85, 056202 (2012), after J. Timmer, Phys. Rev. E 58, 5153 (1998). https://journals.aps.org/pre/abstract/10.1103/PhysRevE.85.056202

TimeseriesSurrogates.randomwalk — Function.

randomwalk(n_steps, x₀)

Linear random walk (AR(1) process with a unit root) from Lucio et al. (2012). This is an example of a nonstationary linear process.

References

Lucio et al., Phys. Rev. E 85, 056202 (2012), after J. Timmer, Phys. Rev. E 58, 5153 (1998). https://journals.aps.org/pre/abstract/10.1103/PhysRevE.85.056202

TimeseriesSurrogates.NSAR2 — Function.

NSAR2(n_steps, x₀, x₁)

Cyclostationary AR(2) process from Lucio et al. (2012).

References

Lucio et al., Phys. Rev. E 85, 056202 (2012), after J. Timmer, Phys. Rev. E 58, 5153 (1998). https://journals.aps.org/pre/abstract/10.1103/PhysRevE.85.056202

TimeseriesSurrogates.AR1 — Function.

AR1(n_steps, x₀, k)

Simple AR(1) model with no static transformation from Lucio et al. (2012).

Equations

The system is given by the following map [1]:

\[x(t+1) = k x(t) + a(t),\]

where $a(t)$ is a draw from a normal distribution with zero mean and unit variance. x₀ sets the initial condition and k is the tunable parameter in the map.

References

Lucio et al., Phys. Rev. E 85, 056202 (2012). https://journals.aps.org/pre/abstract/10.1103/PhysRevE.85.056202