Private types

Private types

Private types in module PassiveTracerFlows:

Private types in module PassiveTracerFlows:

PassiveTracerFlows.TracerAdvDiff.ConstDiffParams — Type.

ConstDiffParams(eta, kap, kaph, nkaph, u, v)
ConstDiffParams(eta, kap, u, v)

Returns the params for constant diffusivity problem with time-varying flow.

PassiveTracerFlows.TracerAdvDiff.ConstDiffSteadyFlowParams — Type.

ConstDiffSteadyFlowParams(eta, kap, kaph, nkaph, u, v, g)
ConstDiffSteadyFlowParams(eta, kap, u, v, g)

Returns the params for constant diffusivity problem with time-steady flow.

PassiveTracerFlows.TracerAdvDiff.Vars — Method.

Vars(g)

Returns the vars for constant diffusivity problem on grid g.