Model Transformations

Model Transformations

POMDPModelTools contains several tools for transforming problems into other classes so that they can be used by different solvers.

Fully Observable POMDP

POMDPModelTools.FullyObservablePOMDP — Type.

FullyObservablePOMDP(mdp)

Turn MDP mdp into a POMDP where the observations are the states of the MDP.

Generative Belief MDP

Every POMDP is an MDP on the belief space GenerativeBeliefMDP creates a generative model for that MDP.

<aside class="warning"> WARNING: The reward generated by the GenerativeBeliefMDP is the reward for a single state sampled from the belief; it is not the expected reward for that belief transition (though, in expectation, they are equivalent of course). Implementing the model with the expected reward requires a custom implementation because belief updaters do not typically deal with reward. </aside>

POMDPModelTools.GenerativeBeliefMDP — Type.

GenerativeBeliefMDP(pomdp, updater)

Create a generative model of the belief MDP corresponding to POMDP pomdp with belief updates performed by updater.

Example

using POMDPModels
using POMDPModelTools
using BeliefUpdaters

pomdp = BabyPOMDP()
updater = DiscreteUpdater(pomdp)

belief_mdp = GenerativeBeliefMDP(pomdp, updater)
@show statetype(belief_mdp) # POMDPModels.BoolDistribution

for (a, r, sp) in stepthrough(belief_mdp, RandomPolicy(belief_mdp), "a,r,sp", max_steps=5)
    @show a, r, sp
end

Underlying MDP

POMDPModelTools.UnderlyingMDP — Type.

UnderlyingMDP(pomdp)

Transform POMDP pomdp into an MDP where the states are fully observed.