Finding axis minima and step sizes

To assign boxes to points, we need a reference point. Throughout this package, the minima along each coordinate axis of the space is used as the reference point.

The following function takes a set of points and a binning scheme ϵ, and returns what the minima along each coordinate axis is, along with the step sizes along each axis resulting from the binning scheme provided by ϵ.

StateSpaceReconstruction.RectangularPartitions.minima_and_stepsizes — Method.

minima_and_stepsizes(points, ϵ) -> (Vector{Float}, Vector{Float})

Find the minima along each axis of the embedding, and computes appropriate stepsizes given ϵ, which provide instructions on how to grid the space. Assumes each point is a column vector.

Specifically, the binning procedure is controlled by the type of ϵ:

ϵ::Int divides each axis into ϵ intervals of the same size.

ϵ::Float divides each axis into intervals of size ϵ.

ϵ::Vector{Int} divides the i-th axis into ϵᵢ intervals of the same size.

ϵ::Vector{Float64} divides the i-th axis into intervals of size ϵᵢ.