Random Policy

Random Policy

A policy that returns a randomly selected action using rand(rng, actions(pomdp)).

POMDPPolicies.RandomPolicy — Type.

RandomPolicy{RNG<:AbstractRNG, P<:Union{POMDP,MDP}, U<:Updater}

a generic policy that uses the actions function to create a list of actions and then randomly samples an action from it.

Constructor:

`RandomPolicy(problem::Union{POMDP,MDP};
         rng=Random.GLOBAL_RNG,
         updater=NothingUpdater())`

Fields

rng::RNG a random number generator

probelm::P the POMDP or MDP problem

updater::U a belief updater (default to NothingUpdater in the above constructor)

POMDPPolicies.RandomSolver — Type.

solver that produces a random policy