Reference Manual

Reference Manual

Twiddle.count_0000_nibbles — Method.

count_0000_nibbles(x::Unsigned)

Counts the number of nibbles (aligned 4 bit segments) in an unsigned integer x that have all their bits unset i.e. nibbles of 0000.

E.g. An input of:

0x0F11F111F11111F1

Would give the answer: 1.

Twiddle.count_00_bitpairs — Method.

count_00_bitpairs(x::Unsigned)

Count the number of bit pairs in an unsigned integer x which have no bits set.

E.g. An input of:

0x0F11F111F11111F1

Would give the answer: 13.

Twiddle.count_01_bitpairs — Method.

count_01_bitpairs(x::Unsigned)

Count the number of bit pairs in an unsigned integer x which have just their LSB set.

E.g. An input of:

0x0F11F111F11111F1

Would give the answer: 11.

Twiddle.count_10_bitpairs — Method.

count_10_bitpairs(x::Unsigned)

Count the number of bit pairs in an unsigned integer x which have just their MSB set.

E.g. An input of:

0x8F11F111F11A11F1

Would give the answer: 3.

Twiddle.count_1111_nibbles — Method.

count_1111_nibbles(x::Unsigned)

Counts the number of nibbles (aligned 4 bit segments) in an unsigned integer x that have all their bits set i.e. counts all nibbles of 1111 in an integer.

E.g. An input of:

0x0F11F111F11111F1

Would give the answer: 4.

Twiddle.count_11_bitpairs — Method.

count_11_bitpairs(x::Unsigned)

Count the number of bit pairs in an unsigned integer x which have both bits set.

E.g. An input of:

0x0F11F111F11111F1

Would give the answer: 8.

Twiddle.count_nonzero_bitpairs — Method.

count_nonzero_bitpairs(x::Unsigned)

Count the number of bit pairs in an unsigned integer x which have at least one bit set.

E.g. An input of:

0x0F11F111F11111F1

Would give the answer: 19.

Twiddle.count_nonzero_nibbles — Method.

count_nonzero_nibbles(x::Unsigned)

Count the number of nibbles (aligned 4 bit segments) in an unsigned integer x which have at least one bit set.

E.g. An input of:

0x0F11F111F11111F1

Would give the answer: 15.

Twiddle.enumerate_nibbles — Method.

enumerate_nibbles{T<:Unsigned}(x::T)

Count the number of set bits in each nibble (aligned 4 bit segments) of an unsigned integer x.

E.g. An input of:

0100 0010 0001 0110 1100 1110 1101 1111

Would result in:

0001 0001 0001 0010 0010 0011 0011 0100

This is used to identify different occurances of certain bit patterns.

Twiddle.mask — Method.

mask{T<:Unsigned}(::Type{T}, n::Integer)

Creates a bit mask for given number of bits n.

The mask starts from the least significant bit, and end at bit n.

e.g:

julia> Twiddle.mask(UInt64, 8)
0x00000000000000ff

Twiddle.nibble_capacity — Method.

nibble_capacity{T<:Unsigned}(::Type{T})

Returns the number of nibbles that an unsigned integer of type T holds. This is essentially twice the size of the type (in bytes).

Twiddle.nibble_mask — Method.

nibble_mask{T<:Unsigned}(value::T, x::T)

Create a mask for the nibbles (aligned 4 bit segments) in an unsigned integer x that filter nibbles matching the corresponding nibble in value.

Twiddle.repeatpattern — Method.

repeatpattern{T<:Unsigned}(::Type{T}, pattern::Unsigned)

Repeats the bitpattern of the pattern integer throughout the an Unsigned integer of type T. Note that this assumed the size of T (in bits) is larger than that of the pattern integer.

This is useful for bit-twiddling code that is to work on any word size.

Let's use a very simple example to show why this is useful. Let's say you have a function in which a byte is masked with a bitmask 0x33.

f(x::UInt8) = x & 0x33

Suppose you wanted this to work for any word size, so 8 bit, 16 bit, 32 bit, 64 bit, and 128 bit words? One way is to write f as a set of seperate methods:

f(x::UInt8) = x & 0x33
f(x::UInt16) = x & 0x3333
f(x::UInt32) = x & 0x33333333
f(x::UInt64) = x & 0x3333333333333333
f(x::UInt128) = x & 0x33333333333333333333333333333333

This is clearly not ideal, you would be writing out the same method multiplet times, but with a differently sized literal for the mask each time.

Alternatively, you could write one parametric function with repeatbyte:

f2(x::T) where {T<:Unsigned} = x & Twiddle.repeatpattern(T, 0x33)
# Or a non-parametric version using `typeof`.
f3(x::Unsigned) = x & Twiddle.repeatpattern(typeof(x), 0x33)

You might expect this to be less efficient - repeatpattern uses several operations to generate the values 0x33, 0x3333 and so on, whereas in the seperate methods, those literal values are hard coded. However, thanks to constant folding during compilation, those operations are done once at compilation time and so the native instructions generated are identical. You can check this with @code_llvm or @code_native.

Twiddle.swapbits — Method.

swapbits{T<:Unsigned}(x::T, i::Integer, j::Integer)

Swap the i'th and j'th bits in an unsigned integer. Note this uses zero based indexes for i and j.

E.g. to swap the LSB and MSB of a byte: 1001 1000 (0x98) -> 0001 1001 (0x19)

swapbits(0x98, 0, 7)