Grids

Grids

UncertainData.InterpolationAndGrids.RegularGrid — Type.

RegularGrid

Fields

min: The minimum value of the grid.

max: The maximum value of the grid.

step: The interval size.

extrapolation_bc: The extrapolation condition. Can also be NaN.

Syntax

Uncertain index-value datasets

The following methods are available for the interpolating of a realization of an uncertain index-value dataset:

No constraints

UncertainData.Resampling.resample — Method.

resample(udata::UncertainIndexValueDataset, 
    grid_indices::InterpolationGrid;
    trunc::TruncateQuantiles = TruncateQuantiles(0.001, 0.999))

Draw a realization of udata, then interpolate the data values to grid_indices.

To avoid very large spans of interpolation, the uncertain indices are truncated to some large quantile range. Values are not truncated.

Sequential constraints

UncertainData.Resampling.resample — Method.

resample(udata::UncertainIndexValueDataset, 
    sequential_constraint::SequentialSamplingConstraint,
    grid_indices::InterpolationGrid;
    trunc::TruncateQuantiles = TruncateQuantiles(0.001, 0.999))

Draw a realization of udata, enforcing a sequential_constraint on the indices. Then, interpolate the values of the realization to the provided grid of indices (grid_indices).

To avoid very large spans of interpolation, the uncertain indices are truncated to some large quantile range. Values are not truncated.