Is a particular constraint applicable?

In addition to the generic sampling constraints, you may impose sequential sampling constraints when resampling an uncertain dataset.

Is a particular constraint applicable?

Not all sequential sampling constraints may be applicable to your dataset. Use these functions to check whether a particular constraint is possible to apply to your dataset.

Syntax

Sequential constraint only

A dataset may be sampling imposing a sequential sampling constraint, but leaving the furnishing distributions untouched otherwise.

UncertainData.Resampling.resample — Method.

resample(udata::AbstractUncertainValueDataset, 
    sequential_constraint::SequentialSamplingConstraint;
    quantiles = [0.0001, 0.9999])

Resample a dataset by imposing a sequential sampling constraint.

Before drawing the realization, all furnishing distributions are truncated to the provided quantiles range. This is to avoid problems in case some distributions have infinite support.

Regular constraint(s) + sequential constraint

Another option is to first impose constraints on the furnishing distributions, then applying the sequential sampling constraint.

UncertainData.Resampling.resample — Method.

resample(udata::AbstractUncertainValueDataset, 
    constraint::Union{SamplingConstraint, Vector{SamplingConstraint}}, 
    sequential_constraint::SequentialSamplingConstraint;
    quantiles = [0.0001, 0.9999])

Resample a dataset by first imposing regular sampling constraints on the furnishing distributions, then applying a sequential sampling constraint.

Before drawing the realization, all furnishing distributions are truncated to the provided quantiles range. This is to avoid problems in case some distributions have infinite support.

List of sequential resampling schemes

StrictlyIncreasing sequences.

StrictlyDecreasing sequences.