-

The following sampling constraints are available:

Generic resampling

These constraints may be used in any resampling setting.

Standard deviation

UncertainData.SamplingConstraints.TruncateStd — Type.

TruncateStd(nσ::Int)

A constraint indicating that the distribution furnishing an uncertain value should be truncated at nσ (n standard deviations).

Value range

UncertainData.SamplingConstraints.TruncateMinimum — Type.

TruncateMinimum(min::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated below at some specified minimum value.

UncertainData.SamplingConstraints.TruncateMaximum — Type.

TruncateMaximum(max::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated above at some specified maximum value.

UncertainData.SamplingConstraints.TruncateRange — Type.

TruncateRange(min::Number, max::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated at some range [min, max].

Quantile range

UncertainData.SamplingConstraints.TruncateLowerQuantile — Type.

TruncateLowerQuantile(lower_quantile::Float64)

A constraint indicating that the distribution furnishing an uncertain value should be truncated below at some quantile.

UncertainData.SamplingConstraints.TruncateUpperQuantile — Type.

TruncateUpperQuantile(upper_quantile::Float64)

A constraint indicating that the distribution furnishing an uncertain value should be truncated above at some quantile.

UncertainData.SamplingConstraints.TruncateQuantiles — Type.

TruncateQuantiles(lower_quantile::Float64, upper_quantile::Float64)

A constraint indicating that the distribution furnishing an uncertain value should be truncated at some quantile quantile (lower_quantile, upper_quantile).