-

There are a few built-in functions to check if your dataset allows the application of certain sequential sampling constraints. These functions will check whether a valid sequence through your dataset exists, so that you can know beforehand whether a particular resampling scheme is possible to apply to your data.

Strictly increasing sequence

UncertainData.SamplingConstraints.strictly_increasing_sequence_exists — Function.

strictly_increasing_sequence_exists(udata::AbstractUncertainValueDataset; 
    quantiles = [0.0001, 0.9999])

Does a path through the dataset exist? I.e, check that a strictly increasing sequence can be found after first constraining each distribution to the provided quantile range (this is necessary because some distributions may have infinite support).

Strictly decreasing sequence

UncertainData.SamplingConstraints.strictly_decreasing_sequence_exists — Function.

strictly_decreasing_sequence_exists(udata::AbstractUncertainValueDataset;
    quantiles = [0.0001, 0.9999])

Does a path through the dataset exist? I.e, check that a strictly decreasing sequence can be found after first constraining each distribution to the provided quantile range (this is necessary because some distributions may have infinite support).