-

The following constraints may be used to impose sequential constraints when sampling a dataset element-wise.

Increasing

UncertainData.SamplingConstraints.StrictlyIncreasing — Type.

StrictlyIncreasing

A sampling constraint indicating element-wise sampling of the uncertain values in a dataset, such that the values of the draw are strictly increasing in magnitude.

Typically used when there are known, physical constraints on the measurements. For example, geochemical measurements of sediments at different depths of a sediment core are taken at physically separate depths in the core. Thus, the order of the indices cannot be flipped, and must be strictly decreasing/increasing.

Decreasing

UncertainData.SamplingConstraints.StrictlyDecreasing — Type.

StrictlyDecreasing

A sampling constraint indicating element-wise sampling of the uncertain values in a dataset, such that the values of the draw are strictly decreasing in magnitude.

Typically used when there are known, physical constraints on the measurements. For example, geochemical measurements of sediments at different depths of a sediment core are taken at physically separate depths in the core. Thus, the order of the indices cannot be flipped, and must be strictly decreasing/increasing.