Abstraction for Statistical Models

Abstraction for Statistical Models

This package defines an abstract type StatisticalModel, and an abstract subtype RegressionModel.

Particularly, instances of StatisticalModel implement the following methods.

StatsBase.adjr2 — Function.

adjr2(obj::StatisticalModel)
adjr²(obj::StatisticalModel)

Adjusted coefficient of determination (adjusted R-squared).

For linear models, the adjusted R² is defined as $1 - (1 - (1-R^2)(n-1)/(n-p))$, with $R^2$ the coefficient of determination, $n$ the number of observations, and $p$ the number of coefficients (including the intercept). This definition is generally known as the Wherry Formula I.

adjr2(obj::StatisticalModel, variant::Symbol)
adjr²(obj::StatisticalModel, variant::Symbol)

Adjusted pseudo-coefficient of determination (adjusted pseudo R-squared).

For nonlinear models, one of the several pseudo R² definitions must be chosen via variant. The only currently supported variant is :MacFadden, defined as $1 - (\log (L) - k)/\log (L0)$. In this formula, $L$ is the likelihood of the model, $L0$ that of the null model (the model including only the intercept). These two quantities are taken to be minus half deviance of the corresponding models. $k$ is the number of consumed degrees of freedom of the model (as returned by dof).

StatsBase.aic — Function.

aic(obj::StatisticalModel)

Akaike's Information Criterion, defined as $-2 \log L + 2k$, with $L$ the likelihood of the model, and k its number of consumed degrees of freedom (as returned by dof).

StatsBase.aicc — Function.

aicc(obj::StatisticalModel)

Corrected Akaike's Information Criterion for small sample sizes (Hurvich and Tsai 1989), defined as $-2 \log L + 2k + 2k(k-1)/(n-k-1)$, with $L$ the likelihood of the model, $k$ its number of consumed degrees of freedom (as returned by dof), and $n$ the number of observations (as returned by nobs).

StatsBase.bic — Function.

bic(obj::StatisticalModel)

Bayesian Information Criterion, defined as $-2 \log L + k \log n$, with $L$ the likelihood of the model, $k$ its number of consumed degrees of freedom (as returned by dof), and $n$ the number of observations (as returned by nobs).

StatsBase.coef — Function.

coef(obj::StatisticalModel)

Return the coefficients of the model.

StatsBase.coefnames — Function.

coefnames(obj::StatisticalModel)

Return the names of the coefficients.

StatsBase.coeftable — Function.

coeftable(obj::StatisticalModel)

Return a table of class CoefTable with coefficients and related statistics.

StatsBase.confint — Function.

confint(obj::StatisticalModel)

Compute confidence intervals for coefficients.

StatsBase.deviance — Function.

deviance(obj::StatisticalModel)

Return the deviance of the model relative to a reference, which is usually when applicable the saturated model. It is equal, up to a constant, to $-2 \log L$, with $L$ the likelihood of the model.

StatsBase.dof — Function.

dof(obj::StatisticalModel)

Return the number of degrees of freedom consumed in the model, including when applicable the intercept and the distribution's dispersion parameter.

StatsBase.fit — Function.

fit(Histogram, data[, weight][, edges]; closed=:left, nbins)

Fit a histogram to data.

Arguments

data: either a vector (for a 1-dimensional histogram), or a tuple of vectors of equal length (for an n-dimensional histogram).

weight: an optional AbstractWeights (of the same length as the data vectors), denoting the weight each observation contributes to the bin. If no weight vector is supplied, each observation has weight 1.

edges: a vector (typically an AbstractRange object), or tuple of vectors, that gives the edges of the bins along each dimension. If no edges are provided, these are determined from the data.

Keyword arguments

closed: if :left (the default), the bin intervals are left-closed [a,b); if :right, intervals are right-closed (a,b].

nbins: if no edges argument is supplied, the approximate number of bins to use along each dimension (can be either a single integer, or a tuple of integers).

Examples

# Univariate
h = fit(Histogram, rand(100))
h = fit(Histogram, rand(100), 0:0.1:1.0)
h = fit(Histogram, rand(100), nbins=10)
h = fit(Histogram, rand(100), weights(rand(100)), 0:0.1:1.0)
h = fit(Histogram, [20], 0:20:100)
h = fit(Histogram, [20], 0:20:100, closed=:right)

# Multivariate
h = fit(Histogram, (rand(100),rand(100)))
h = fit(Histogram, (rand(100),rand(100)),nbins=10)

Fit a statistical model.

StatsBase.fit! — Function.

Fit a statistical model in-place.

StatsBase.informationmatrix — Function.

informationmatrix(model::StatisticalModel; expected::Bool = true)

Return the information matrix. By default the Fisher information matrix is returned, while the observed information matrix can be requested with expected = false.

StatsBase.isfitted — Function.

isfitted(obj::StatisticalModel)

Indicate whether the model has been fitted.

StatsBase.islinear — Function.

islinear(obj::StatisticalModel)

Indicate whether the model is linear.

StatsBase.loglikelihood — Function.

loglikelihood(obj::StatisticalModel)

Return the log-likelihood of the model.

StatsBase.mss — Function.

mss(obj::StatisticalModel)

Return the model sum of squares.

StatsBase.nobs — Function.

nobs(obj::StatisticalModel)

Return the number of independent observations on which the model was fitted. Be careful when using this information, as the definition of an independent observation may vary depending on the model, on the format used to pass the data, on the sampling plan (if specified), etc.

StatsBase.nulldeviance — Function.

nulldeviance(obj::StatisticalModel)

Return the deviance of the null model, that is the one including only the intercept.

StatsBase.r2 — Function.

r2(obj::StatisticalModel)
r²(obj::StatisticalModel)

Coefficient of determination (R-squared).

For a linear model, the R² is defined as $ESS/TSS$, with $ESS$ the explained sum of squares and $TSS$ the total sum of squares.

StatsBase.rss — Function.

rss(obj::StatisticalModel)

Return the residual sum of squares.

StatsBase.score — Function.

score(obj::StatisticalModel)

Return the score of the statistical model. The score is the gradient of the log-likelihood with respect to the coefficients.

StatsBase.stderror — Function.

stderror(obj::StatisticalModel)

Return the standard errors for the coefficients of the model.

StatsBase.vcov — Function.

vcov(obj::StatisticalModel)

Return the variance-covariance matrix for the coefficients of the model.

StatsBase.weights — Method.

weights(obj::StatisticalModel)

Return the weights used in the model.

RegressionModel extends StatisticalModel by implementing the following additional methods.

StatsBase.dof_residual — Function.

dof_residual(obj::RegressionModel)

Return the residual degrees of freedom of the model.

StatsBase.fitted — Function.

fitted(obj::RegressionModel)

Return the fitted values of the model.

StatsBase.leverage — Function.

leverage(obj::RegressionModel)

Return the diagonal of the projection matrix.

StatsBase.meanresponse — Function.

meanresponse(obj::RegressionModel)

Return the mean of the response.

StatsBase.modelmatrix — Function.

modelmatrix(obj::RegressionModel)

Return the model matrix (a.k.a. the design matrix).

StatsBase.response — Function.

response(obj::RegressionModel)

Return the model response (a.k.a. the dependent variable).

StatsBase.predict — Function.

predict(obj::RegressionModel, [newX])

Form the predicted response of model obj. An object with new covariate values newX can be supplied, which should have the same type and structure as that used to fit obj; e.g. for a GLM it would generally be a DataFrame with the same variable names as the original predictors.

StatsBase.predict! — Function.

predict!

In-place version of predict.

StatsBase.residuals — Function.

residuals(obj::RegressionModel)

Return the residuals of the model.