-

Pooled test

UncertainData.UncertainStatistics.ApproximateTwoSampleKSTestPooled — Function.

ApproximateTwoSampleKSTestPooled(d1::UncertainDataset,
    d2::UncertainDataset, n::Int = 1000) -> ApproximateTwoSampleKSTest

First, draw n realisations of each uncertain value in d1, then separately draw n realisations of each uncertain value in d2. Then, pool all realisations for d1 together and all realisations of d2 together.

On the pooled realisations, perform an asymptotic two-sample Kolmogorov–Smirnov-test of the null hypothesis that the distribution furnishing the d1 value pool represents the same distribution as the distribution furnishing the d2 value pool, against the alternative hypothesis that the furnishing distributions are different.

Element-wise test

UncertainData.UncertainStatistics.ApproximateTwoSampleKSTestElementWise — Function.

ApproximateTwoSampleKSTestElementWise(d1::UncertainDataset,
    d2::UncertainDataset, n::Int = 1000) -> Vector{ApproximateTwoSampleKSTest}

Assuming d1 and d2 contain the same number of uncertain observations, draw n realisations of each uncertain value in d1, then separately and separately draw n realisations of each uncertain value in d2.

Then, perform an asymptotic two-sample Kolmogorov–Smirnov-test of the null hypothesis that the uncertain values in d1 and d2 come from the same distribution against the alternative hypothesis that the (element-wise) values in d1 and d2 come from different distributions.

The test is performed pairwise, i.e. ApproximateTwoSampleKSTest(d1[i], d2[i]) with n draws for the $i$-ith pair of uncertain values.