Documentation

The UncertainScalarPopulation type allows representation of an uncertain scalar represented by a population of values who will be sampled according to a vector of explicitly provided probabilities. Think of it as an explicit kernel density estimate.

Documentation

UncertainData.UncertainValues.UncertainScalarPopulation — Type.

UncertainScalarPopulation

An uncertain value represented by a population for which the probabilities of the values are dictated by a set of weights. The weights are normalized by default.

Examples

The two following ways of constructing populations are equivalent.

values = rand(1:50, 100)
weights = rand(100)
population = UncertainValue(values, weights)

values = rand(1:50, 100)
weights = rand(100)
population = UncertainScalarPopulation(values, weights)