Index

MOSEK Solver API Reference

Create and destroy MOSEK objects

Mosek.makeenv — Function.

makeenv()

Create a MOSEK environment.

makeenv(func::Function)

Create a MOSEK environment for use with do-syntax.

Mosek.maketask — Function.

maketask(env:Env)
maketask()

Create a task that belongs to either the given env or to the global env.

maketask(task::Task)

Create a clone of task.

maketask(func::Function, env::Env)
maketask(func::Function)

Create a task that either longs to the given env or to the global env. The func parameter is a function

func(task :: Task) :: Any

This can be used with the do-syntax:

maketask(env) do task
    readdata(task,"MyFile.task")
end

Mosek.deleteenv — Function.

deleteenv(t::Env)

Destroy the task object.

Mosek.deletetask — Function.

deletetask(t::Task)

Destroy the task object.

Function reference