-

PowerDynBase.no_internal_differentials — Constant.

A variable to be used when no internal differentials are present for a node dynamics type.

PowerDynBase.no_internal_masses — Constant.

A variable to be used when no internal masses are present for a node dynamics type.

PowerDynBase.AbstractNodeParameters — Type.

Abstract super type for all node parameter types.

PowerDynBase.FourthEq — Type.

FourthEq(H, P, D, Ω, E_f, T_d_dash ,T_q_dash ,X_q_dash ,X_d_dash,X_d, X_q)

A node type that applies the 4th-order synchronous machine model with frequency/angle and voltage dynamics.

Additionally to $u$, it has the internal dynamic variables

$\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$ and

$\theta$ representing the relative angle of the rotor with respect to the voltage angle $\phi$.

Keyword Arguments

H: inertia

P: active (real) power output

D: damping coefficient

Ω: rated frequency of the power grid, often 50Hz

T_d_dash: time constant of d-axis

T_q_dash: time constant of q-axis

X_d_dash: transient reactance of d-axis

X_q_dash: transient reactance of q-axis

X_d: reactance of d-axis

X_d: reactance of q-axis

Mathematical Representation

Using FourthEq for node $a$ applies the equations

\[ u = -je_c e^{j\theta} = -j(e_d + je_q)e^{j\theta}\\ e_c= e_d + je_q = jue^{-j\theta}\\ i = -ji'e^{j\theta} = -j(i_d+ j i_q )e^{j\theta} = Y^L \cdot u \\ i_c= i_d + ji_q = jie^{-j\theta}\\ p = \Re (i^* u)\]

The fourth-order equations read (according to Sauer, p. 140, eqs. (6110)-(6114)) and p. 35 eqs(3.90)-(3.91)

\[ \frac{d\theta}{dt} = \omega \\ \frac{d\omega}{dt} = P-D\omega - p -(x'_q-x'_d)i_d i_q\\ \frac{d e_q}{dt} = \frac{1}{T'_d} (- e_q - (x_d - x'_d) i_{d}+ e_f) \\ \frac{d e_d}{dt} = \frac{1}{T'_q} (- e_d + (x_q - x'_q) i_{q}) \\\]

With the PowerDynamics.jl \time{naming conventions} of $i$ and $u$ they read as

\[ \dot u = \frac{d}{dt}(-j e_c e^{j\theta})=-j(\dot e_d + j\dot e_q)e^{j\theta} + uj\omega\]

source

PowerDynBase.GridDynamics — Type.

Abstract super type for all abstract grid dynamics types.

PowerDynBase.GridDynamics — Method.

function GridDynamics(nodes::AbstractVector{<:AbstractNodeParameters}, args...; kwargs...)

Convert all subtypes of PowerDynBase.AbstractNodeParameters to the corresponding subtypes of PowerDynBase.AbstractNodeDynamics and then call PowerDynBase.GridDynamics again.

PowerDynBase.GridDynamics — Method.

function GridDynamics(
    nodes::AbstractArray{<:AbstractNodeDynamics},
    LY::AbstractMatrix;
    skip_LY_check=false,
    kwargs...)

Bring all sutypes of PowerDynBase.AbstractNodeDynamics on one level and then create for each subtype of PowerDynBase.AbstractNodeDynamics the corresponding subtype of PowerDynBase.GridDynamics by using PowerDynBase._GridDynamics.

PowerDynBase.GridDynamicsError — Type.

Error to be thrown if something goes wrong during the grid dynamics construction.

PowerDynBase.NetworkRHS — Type.

struct NetworkRHS{T, M} <: AbstractNetworkFunction{T, M}
    nodes::AbstractVector{T}
    LY::M
    numnodes
    systemsize
    intrange # unitrange telling me where I find the internal dynamic variables
    nodalintranges # unit ranges to find the internal variables for each node in the full length of internal variables
end

Representing the full dynamics of the power grid.

source

PowerDynBase.NetworkRHS — Method.

NetworkRHS(x)

Return the struct of type PowerDynBase.NetworkRHS for x.

PowerDynBase.NetworkRHS — Method.

(rhs::NetworkRHS)(x::AbstractDEVariable, t)

Evalate the network right-hand-side function.

PowerDynBase.NetworkRHS — Method.

NetworkRHS(nodes::AbstractVector{T}, LY::M) where {T<:AbstractNodeDynamics, M<:AbstractMatrix}

Create an PowerDynBase.NetworkRHS object from a node list and the nodal admittance matrix.

PowerDynBase.NodeDynamicsError — Type.

Error to be thrown if something goes wrong during the node dynamics construction.

PowerDynBase.OrdinaryNodeDynamics — Type.

OrdinaryNodeDynamics(;rhs, n_int)

The type representing the dynamics of a node that is described via ODEs.

Each node $a$ has the complex voltage $u$ and $n$ real internal variables $y_1, \dots, y_n$, so it generally describes a system of ordinary differential equation as

\[\frac{du_a}{dt} = f_u(u_a, {i_c}_a, y_1, \dots, y_n) \\ \frac{dy_{ak}}{dt} = f_k(u_a, {i_c}_a, y_1, \dots, y_n)\quad \forall k = 1, \dots, n.\]

$f$ is represented by rhs field of OrdinaryNodeDynamics.

the general signature of rhs is

rhs(dint_dt::AbstractVector,
    u::Complex,
    i::Complex,
    int::AbstractVector,
    t,
    )::Complex

Input

u is the complex voltage $u$

i is the complex current $i$

int is the array of internal variables $y_1, \dots, y_n$

t is the time $t$

Output

the (complex) return value describes $\frac{du}{dt}$

rhs writes values in dint_dt describing the left-hand side $\frac{dy_1}{dt}, \dots, \frac{dy_n}{dt}$

source

PowerDynBase.OrdinaryNodeDynamicsWithMass — Type.

OrdinaryNodeDynamicsWithMass(;rhs, n_int, m_u, m_int)

The type representing the dynamics of a node that is described via ODEs.

Each node $a$ has the complex voltage $u$ and $n$ (= n_int) real internal variables $y_1, \dots, y_n$, so it generally describes a system of ordinary differential equation with a voltage mass $m_u$ and internal masses $m^{int}_1, \dots, m^{int}_n$ as

\[m_u\frac{du_a}{dt} = f_u(u_a, {i_c}_a, y_1, \dots, y_n) \\ m^{int}_k\frac{dy_{ak}}{dt} = f_k(u_a, {i_c}_a, y_1, \dots, y_n)\quad \forall k = 1, \dots, n.\]

As we assume that all masses are binary (either 1, or 0), that means, one can implement semi-explicit differential algebraic equations with this node dynamics type. $f$ is represented by rhs field of OrdinaryNodeDynamics.

the general signature of rhs is

rhs(dint_dt::AbstractVector,
    u::Complex,
    i::Complex,
    int::AbstractVector,
    t,
    )::Complex

Input

u is the complex voltage $u$

i is the complex current $i$

int is the array of internal variables $y_1, \dots, y_n$

t is the time $t$

Output

the (complex) return value describes $\frac{du}{dt}$

rhs writes values in dint_dt describing the left-hand side $\frac{dy_1}{dt}, \dots, \frac{dy_n}{dt}$

The binary masses are:

m_u is the boolean value for $m_u$

m_int is the array of boolean values for $m^{int}_1, \dots, m^{int}_n$

PowerDynBase.PQAlgebraic — Type.

PQAlgebraic(;S)

A node type that locally fixes the active ($P$) and reactive power ($Q$) output of the node.

Keyword Arguments

S = P + Q*im: the complex power output

Mathematical Representation

Using PQAlgebraic for node $a$ applies the equation

\[0 = S_a - u_a \cdot i_a^*.\]

source

PowerDynBase.PVAlgebraic — Type.

PVAlgebraic(;P,V)

A node type that locally fixes the active power ($P$) and the voltage magnitude ($V$) of the node.

Keyword Arguments

P: the active (real) power output

V: voltage magnitude

Mathematical Representation

Using PVAlgebraic for node $a$ applies the equations

\[0 = P_a - \Re\left(u_a \cdot i_a^*\right), \\ 0 = V_a - \left\|u_a\right\|.\]

source

PowerDynBase.SlackAlgebraic — Type.

SlackAlgebraic(;U)

A node type that locally fixes the complex voltage ($U$) of the node.

As the complex voltage can be represented as $U=Ve^{i\phi}$, this is equivlant to fixing the voltage magnitude $V$ and the angle $\phi$.

Keyword Arguments

U: the complex voltage

Mathematical Representation

Using SlackAlgebraic for node $a$ applies the equation

\[0 = U_a - u_a.\]

source

PowerDynBase.State — Type.

    State(base; t=nothing)
    State(grid, vec; t=nothing)

Encode the information on the value of a state vector at a particular time point.

Keyword Arguments

base is an instance of a BaseState, essentially it contains the state vector and the complete rhs of the system. Instead of base, you can also directly use a GridDynamics instance grid and a properly sized state vector vec to instantiate a State.

t is a time point associated to the base. It defaults to nothing.

Indexing

Concerning the indexing, a State object $s$ basically behaves like a an array. There are plenty of convenient ways to access its contents at a node $j$ by using a particular symbol:

s[j, :u]: complex voltage

s[j, :v]: voltage magnitude

s[j, :φ]: voltage angle

s[j, :i]: complex nodal current

s[j, :iabs]: nodal current magnitude

s[j, :δ]: nodal current angle

s[j, :s]: apparent power

s[j, :p]: real power

s[j, :q]: reactive power

Currently, setting the state value is only implemented for $u$ and $v$, the other quantities are derived automatically.

When a node $j$ has internal variables, you can access (and set) the $k$-th internal variable by calling

s[j, :int, k].

The internal variables can be also directly accessed with symbols, i.e.

s[j, :ω]

returns the frequency $ω$ at node $j$. To find out the proper symbol, the easiest way is to look into the docs of the corresponding AbstractNodeParameters subtype, check the output of internalsymbolsof or simply look at the output of println:

julia> internalsymbolsof(SwingEq(H=2, P=3, D=4, Ω=5))
1-element Array{Symbol,1}:
 :ω

julia> println(SwingEq(H=2, P=3, D=4, Ω=5))
SwingEq[:ω](H=2, P=3, D=4, Ω=5)

PowerDynBase.StateError — Type.

Error to be thrown if something goes wrong when creating or modifying states.

PowerDynBase.SwingEq — Type.

SwingEq(;H, P, D, Ω)

A node type that applies the swing equation to the frequency/angle dynamics and keeps the voltage magnitude as is.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

H: inertia

P: active (real) power output

D: damping coefficient

Ω: rated frequency of the power grid, often 50Hz

Mathematical Representation

Using SwingEq for node $a$ applies the equations

\[\frac{du_a}{dt} = i u_a \omega_a, \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

which is equivalent to

\[\frac{d\phi_a}{dt} = \omega, \\ v = v(t=0) = \text{const.} \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

source

PowerDynBase.SwingEqLVS — Type.

SwingEqLVS(;H, P, D, Ω, Γ, V)

A node type that applies the swing equation to the frequency/angle dynamics and has a linear voltage stability (LVS) term.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

H: inertia

P: active (real) power output

D: damping coefficient

Ω: rated frequency of the power grid, often 50Hz

Γ: voltage stability coefficient

V: set voltage, usually 1

Mathematical Representation

Using SwingEq for node $a$ applies the equations

\[\frac{du_a}{dt} = i u_a \omega - \frac{u}{\|u\|} Γ_a (v_a - V_a), \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

which is equivalent to

\[\frac{d\phi_a}{dt} = \omega_a, \\ \frac{dv_a}{dt} = - Γ_a (v_a - V_a) \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

source

PowerDynBase.VSIMinimal — Type.

VSIMinimal(;τ_P,τ_Q,K_P,K_Q,E_r,P,Q)

A node type that applies the frequency and voltage droop control to control the frequency and voltage dynamics.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

τ_p: time constant active power measurement

τ_Q: time constant reactive power measurement

K_P: droop constant frequency droop

K_Q: droop constant voltage droop

V_r: reference/ desired voltage

P: active (real) power infeed

Q: reactive (imag) power infeed

Mathematical Representation

Using VSIMinimal for node $a$ applies the equations

\[\dot{\phi}_a=\omega_a\\ \dot{\omega}_a=\frac{1}{\tau_{P,a}}[-\omega_a-K_{P,a} (\Re\left(u_a \cdot i_a^*\right)-P_{ref,a})]\\ \tau_Q\dot{v}_a=-v_a+V_{ref}-K_{Q,a} (\Im\left(u_a \cdot i_a^*\right)-Q_{ref,a})\\ \dot{u}_a=\dot{v_a}e^{j\phi}+j\omega_a u_a\]

```

source

PowerDynBase.VSIVoltagePT1 — Type.

VSIVoltagePT1(;τ_v,τ_P,τ_Q,K_P,K_Q,E_r,P,Q)

A node type that applies the frequency and voltage droop control to control the frequency and voltage dynamics.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

τ_v: time constant voltage control delay

τ_p: time constant active power measurement

τ_Q: time constant reactive power measurement

K_P: droop constant frequency droop

K_Q: droop constant voltage droop

V_r: reference/ desired voltage

P: active (real) power infeed

Q: reactive (imag) power infeed

Mathematical Representation

Using VSIVoltagePT1 for node $a$ applies the equations

\[\dot{\phi}_a=\omega_a\\ \dot{\omega}_a=\frac{1}{\tau_{P,a}}[-\omega_a-K_{P,a} (\Re\left(u_a \cdot i_a^*\right)-P_{ref,a})]\\ \tau_v\dot{v}_{a}=-v_a+V_{ref}-K_{Q,a}(q_{m,a}-Q_{ref,a})\\ \tau_Q \dot{q}_{m,a}=-q_{m,a}+\Im\left(u_a \cdot i_a^*\right)\\ \dot{u}_a=\dot{v_a}e^{j\phi}+j\omega_a u_a\\\]

source

Base.convert — Method.

convert(::Type{OrdinaryNodeDynamicsWithMass}, ::OrdinaryNodeDynamics)

Conversion of OrdinaryNodeDynamics to OrdinaryNodeDynamicsWithMass by assuming all masses are 1 (true).

Base.convert — Method.

convert(::Type{AlgebraicNodeDynamics}, ::OrdinaryNodeDynamicsWithMass)

Conversion of OrdinaryNodeDynamicsWithMass to AlgebraicNodeDynamics by transforming a RHS function into a root function.

Base.convert — Method.

convert(::Type{AlgebraicNodeDynamics}, ::OrdinaryNodeDynamics)

Conversion of OrdinaryNodeDynamics to AlgebraicNodeDynamics by going via OrdinaryNodeDynamicsWithMass.

Base.promote_rule — Method.

promote_rule(::Type{OrdinaryNodeDynamics}, ::Type{OrdinaryNodeDynamicsWithMass}) = OrdinaryNodeDynamicsWithMass

OrdinaryNodeDynamics can be promoted to OrdinaryNodeDynamicsWithMass, see PowerDynBase.convert.

Base.promote_rule — Method.

promote_rule(::Type{OrdinaryNodeDynamicsWithMass}, ::Type{AlgebraicNodeDynamics}) = AlgebraicNodeDynamics

OrdinaryNodeDynamicsWithMass can be promoted to AlgebraicNodeDynamics, see PowerDynBase.convert.

Base.promote_rule — Method.

promote_rule(::Type{OrdinaryNodeDynamics}, ::Type{AlgebraicNodeDynamics}) = AlgebraicNodeDynamics

OrdinaryNodeDynamics can be promoted to AlgebraicNodeDynamics, see PowerDynBase.convert.

PowerDynBase.AdmittanceLaplacian — Method.

AdmittanceLaplacian(x)

Return the nodal admittance matrix of the system.

PowerDynBase.Nodes — Method.

Nodes(x)

Return the array of nodes for x.

PowerDynBase.SystemSize — Method.

SystemSize(x)

Return the full system size, i.e. number of independent, dynamic, real-valued variables, for x.

PowerDynBase.differentials — Method.

Collect the differnetials (or masses) for the whole grid dynamics from each node dynamics.

PowerDynBase.internaldsymbolsof — Method.

Get the symbols representing the derivative of the internal variables of the node.

PowerDynBase.internaloutsymbolsof — Method.

Get the symbols representing the output of the internal variables of the node.

PowerDynBase.internalsymbolsof — Method.

Get the symbols representing the internal variables of the node.

PowerDynBase.parametersof — Method.

Get the parameters struct for the node.

PowerDynBase.showdefinition — Method.

Show the definition generated by a macro of PowerDynamics.jl, e.g. the macro PowerDynBase.@DynamicNode creating a subtype of PowerDynBase.AbstractNodeParameters.

PowerDynBase.@DynamicNode — Macro.

Macro for creating a new type of dynmic nodes.

Syntax Description:

@DynamicNode MyNewNodeName(Par1, Par2, ...) <: NodeDynamicsType(N1, N2, ...) begin
    [all prepratory things that need to be run just once]
end [[x1, dx1], [x2, dx2]] begin
    [the actual dynamics equation]
    [important to set the output variables]
end

where MyNewNodeName is the name of the new type of dynamic node, Par1, Par2, ... are the names of the parameters, NodeDynamicsType the the node dynamics type (e.g. OrdinaryNodeDynamics or OrdinaryNodeDynamicsWithMass), N1, N1, ... the parameters of the dynamics type, x1, x2, ... the internal variables of the node and dx1, dx2, ... the corresponding differentials.

In the first block, the preparation code that needs to be run only once is inserted. Finally, the second block contains the dynamics description, where it's important that the output variables need to be set. In case of OrdinaryNodeDynamics and OrdinaryNodeDynamicsWithMass, these are du and the differentials of the internal variables (here dx1, dx2).

Below are two examples:

@DynamicNode SwingEqParameters(H, P, D, Ω) <: OrdinaryNodeDynamics() begin
    @assert D > 0 "damping (D) should be >0"
    @assert H > 0 "inertia (H) should be >0"
    Ω_H = Ω * 2pi / H
end [[ω, dω]] begin
    p = real(u * conj(i_c))
    dϕ = ω # dϕ is only a temp variable that Julia should optimize out
    du = u * im * dϕ
    dω = (P - D*ω - p)*Ω_H
end

@DynamicNode SlackAlgebraicParameters(U) <: OrdinaryNodeDynamicsWithMass(m_u=false, m_int=no_internal_masses) begin
end [] begin
        du = u - U
end

PowerDynBase.AbstractAlgebraicGridDynamics — Type.

Abstract super type for all grid dynamics represented by DAEs.

PowerDynBase.AbstractAlgebraicNodeDynamics — Type.

Abstract super type for all node dynamics represented by DAEs.

PowerDynBase.AbstractDAEVariable — Type.

Abstract super type for all Variables for DAE-type node dynamics.

PowerDynBase.AbstractDEVariable — Type.

abstract type AbstractDEVariable end

Abstract super type for all variables that PowerDynBase.AbstractNetworkFunction sub-types can be called with.

DEVariable stands for Differential Equation Variable.

The basic idea of a DEVariable is to combine (the arrays of) all necessary variables for a sub-type of PowerDynBase.GridDynamics. E.g. for a PowerDynBase.OrdinaryGridDynamics one needs only the state variable (here called val for value) and the derivative (here called ddt). Hence, the definition for the corresponding PowerDynBase.ODEVariable is:

@DEVariable struct ODEVariable{Tval, Tddt} <: AbstractODEVariable
    val::AbstractVector{Tval}
    ddt::AbstractVector{Tddt}
end ddt

The final ddt statement is part of the PowerDynBase.@DEVariable macro stating that the output variable of this kind of DEVariable is ddt. Further, the PowerDynBase.@DEVariable macro generates all the necessary constructors. Check its docs for more details.

PowerDynBase.AbstractNetworkFunction — Type.

abstract type AbstractNetworkFunction{T<:AbstractNodeDynamics, M<:AbstractMatrix} end

Abstract super type of all functions that define how a differential equation for the whole network / power grid behaves, e.g. the full right-hand-side function of the ODE.

PowerDynBase.AbstractNodeDynamics — Type.

Abstract super type for all abstract node dynamics types.

PowerDynBase.AbstractODEVariable — Type.

Abstract super type for all Variables for ODE-type node dynamics.

PowerDynBase.AbstractOrdinaryGridDynamics — Type.

Abstract super type for all grid dynamics represented by ODEs.

PowerDynBase.AbstractOrdinaryNodeDynamics — Type.

Abstract super type for all node dynamics represented by ODEs.

PowerDynBase.AlgebraicGridDynamics — Type.

struct AlgebraicGridDynamics <: AbstractAlgebraicGridDynamics
    rhs::NetworkRHS
    differentials::AbstractVector{Bool} # boolean values whether there a variable is a differential
end

The data structure that contains all the information necessary for a power grid model that can be described as an differential algebraic equation. rhs is the PowerDynBase.NetworkRHS. differentials is a 1-dimensional array of boolean values. A true entry means the corresponding variable is dynamic and has a derivative variable. A false entry means the corresponding variable is defined by an algebraic constraint only.

source

PowerDynBase.AlgebraicNodeDynamics — Type.

DOCS TBD!

source

PowerDynBase.BaseState — Type.

    BaseState(grid, vec)

Encode a state vector and the corresponding rhs information.

Keyword Arguments

grid is a GridDynamics instance that contains the overall system rhs.

vec is a state vector of the system who's length is given by the total number of internal and voltage variables.

Indexing

In an instance b of of a BaseState behaves like an array, i.e. you can access the $j$-th element of the state vector (and set it to a value $ξ$) by calling b[j] ( = ξ ).

PowerDynBase.DAEVariable — Type.

Variables for DAE-type node dynamics.

source

PowerDynBase.ODEVariable — Type.

Variables for ODE-type node dynamics.

source

PowerDynBase.OrdinaryGridDynamics — Type.

struct OrdinaryGridDynamics <: AbstractOrdinaryGridDynamics
    rhs::NetworkRHS
end

The data structure that contains all the information necessary for a power grid model that can be described as an ordinary differential equation. In this case, only the PowerDynBase.NetworkRHS is necessary.

source

PowerDynBase.OrdinaryGridDynamicsWithMass — Type.

struct OrdinaryGridDynamicsWithMass <: AbstractAlgebraicGridDynamics
    rhs::NetworkRHS
    masses::AbstractVector{Bool} # diagonal part of the mass matrix, off-diagonal is assumed to be 0 anyway
end

The data structure that contains all the information necessary for a power grid model that can be described as an ordinary differential equation with masses, i.e. a semi-explicit differential algebraic equation. rhs is the PowerDynBase.NetworkRHS. masses is a 1-dimensional array representing the diagonal entries of the mass matrix. The off-diagonal entries are assumed to be 0. masses can only contain boolean values representing: true the equation is treated as a ordinary differential eqation and false the equation is treated as an algebraic constraint on the state variables.

source

PowerDynBase.PowerDynamicsError — Type.

Abstract super type of all PowerDynamics.jl Errors.

Base.view — Method.

Extend view from arrays to subtypes of AbstractDEVariable.

PowerDynBase.DynamicNode — Method.

See PowerDynBase.@DynamicNode.

PowerDynBase._GridDynamics — Method.

Create for each subtype of PowerDynBase.AbstractNodeDynamics the corresponding subtype of PowerDynBase.GridDynamics.

PowerDynBase.checkLY — Method.

Check whether the admittance laplacian has no purely nodal admittances, i.e. that the sum of columns and rows equals to zero.

PowerDynBase.complexview — Method.

Extend PowerDynBase.complexview from arrays to subtypes of AbstractDEVariable.

PowerDynBase.complexview — Method.

complexview(vec::AbstractArray, i0, n)

Interpret (part of) an array of real values as an array with complex values. i0 is the index where to start. n is the number of complex values that should be extracted.

PowerDynBase.create_DEVariable — Method.

Basically, this macro generates all the constructors (internal and external) for a subtype of AbstracDEVariable.

If you want to understand what this macro does, call PowerDynBase.showdefinition with a type that was generated by this macro, e.g. showdefinition(ODEVariable). It will output the full defintion that the macro actually creates.

PowerDynBase.excomparison — Method.

Create an expresseion where == is applied between all the expressions given as argument here.

PowerDynBase.getDEVariableType — Method.

Identify each subtype of AbstractNodeDynamics with its corresponding subtype of AbstractDEVariable

PowerDynBase.int_differentials — Method.

Return the 1-dimensional differentials array (see PowerDynBase.AlgebraicGridDynamics) for the internal variables for each node.

PowerDynBase.internal_unitranges — Method.

Get the unit ranges that indicate where in the array the internal variables for each of the nodes is saved.

PowerDynBase.mapfields — Method.

function mapfields(f, s, args...)

Applies f to all fields of (the struct) s giving args... as additional arguments.

It's written as a @generated function in order to ensure that the compiler can infer the types.

PowerDynBase.nint — Method.

Get number of internal arguments of the node.

PowerDynBase.nodeiterator — Method.

nodeiterator(rhs::NetworkRHS, x::AbstractDEVariable, t)

Distribute the values in x over all the nodes that are summarized in rhs.

PowerDynBase.rhs2root — Method.

A function converting a rhs-type function to a root-type function.

PowerDynBase.symbolsof — Method.

Get the symbols data structure for the node.

PowerDynBase.total_nint — Method.

Get the total number of internal variables for an array of node dynamics.

PowerDynBase.total_nvars — Method.

Get the total number of dynamic variables for an array of node dynamics.

This is basically the (real) dimension of the system, hence the sum of internal dynamic variables + 2*(number of nodes = number of complex voltages). The 2 is due to the fact that the complex voltages are treated as two real variables.

PowerDynBase.u_differentials — Method.

Return the 1-dimensional differentials array (see PowerDynBase.AlgebraicGridDynamics) for the voltagee variables for each node.

PowerDynBase.@DEVariable — Macro.

@DEVariable struct DEVariableName{Type1, Type2, ...} <: AbstractODEVariable
    ...
end outputVariableName

Basically, this macro generates all the constructors (internal and external) for a subtype of AbstracDEVariable. In particular, it creates an external constructor that automatically chooses the type for the outputVariableName and instantiates this variable.

If you want to understand what this macro does, call PowerDynBase.showdefinition with a type that was generated by this macro, e.g. showdefinition(ODEVariable). It will output the full defintion that the macro actually creates.

Furthermore, the macro is just the interface for the PowerDynBase.create_DEVariable function.

PowerDynSolve.GridProblem — Type.

struct GridProblem{P<:DEProblem, S<:AbstractState, T<:AbstractFloat} # T is for the timespan
    prob::P
    start::S
    timespan::Tuple{T, T}
end

Define an analogous problem to DifferentialEquations.jl's DEProblem subtypes. In the future, this is planned to be replaced by a real DEProblem subtype.

PowerDynSolve.GridProblem — Method.

function GridProblem(g::GridDynamics, start::AbstractVector, timespan; kwargs...)

Construct a PowerDynSolve.GridProblem from a an initial condition start for the grid dynamics g which is a subtype of PowerDynBase.GridDynamics.

PowerDynSolve.GridProblem — Method.

function GridProblem(start::AbstractState{OrdinaryGridDynamics, V, T}, timespan; kwargs...) where {V,T}

Construct a PowerDynSolve.GridProblem from a an initial condition (i.e. a PowerDynBase.State) for an PowerDynBase.OrdinaryGridDynamics.

PowerDynSolve.GridProblem — Method.

function GridProblem(start::AbstractState{OrdinaryGridDynamicsWithMass, V, T}, timespan; kwargs...) where {V,T}

Construct a PowerDynSolve.GridProblem from a an initial condition (i.e. a PowerDynBase.State) for an PowerDynBase.OrdinaryGridDynamicsWithMass.

PowerDynSolve.GridProblem — Method.

function GridProblem(g::G, start::AbstractState{G, V, T}, timespan; kwargs...) where {G, V, T}

Construct a PowerDynSolve.GridProblem from a an initial condition start (i.e. a PowerDynBase.State) with the corresponding subtype of PowerDynBase.GridDynamics.

DiffEqBase.solve — Method.

function solve(g::GridDynamics, x0, timespan)

Solve a power grid g (of type PowerDynSolve.GridDynamics) starting at x0 for a timespan, using DifferentialEquations.jl in the back. The correct solvers are automatically chosen.

DiffEqBase.solve — Method.

function solve(p::GridProblem; kwargs...)

Solve a PowerDynSolve.GridProblem, using DifferentialEquations.jl in the back. The correct solvers are automatically chosen.

kwargs are the keyword arguments that are simply passed to DifferentialEquations.jl.

PowerDynSolve.AbstractGridSolution — Type.

Abstract super type for all data structures representing solutions of power grid models.

PowerDynSolve.GridSolution — Type.

struct GridSolution <: AbstractGridSolution
    dqsol::AbstractTimeseriesSolution
    griddynamics::GridDynamics
end

The data structure interfacing to the solution of the differntial equations of a power grid. Normally, it is not created by hand but return from PowerDynSolve.solve.

Accessing the solution in a similar interface as PowerDynBase.State.

For some grid solution sol, one can access the variables as

sol(t, n, s)

where t is the time (either float or array), n the node number(s) (either integer, array, range (e.g. 2:3) or colon (:, for all nodes)), and s is the symbol represnting the chosen value. s can be either: :v, :φ, :i, :iabs, :δ, :s, :p, :q, or the symbol of the internal variable of the nodes. The meaning of the symbols derives from the conventions of PowerDynamics.jl.

Finally, one can access the a-th internal variable of a node by using sol(t, n, :int, a).

Interfacing the Plots.jl library via plotting recipes, that follow similar instructions as the direct access to the solution.

For some grid solution sol, one plot variables of the solution asin

using Plots
plot(sol, n, s, plots_kwargs...)

where n and s are as in the accessing of plots, and plots_kwargs are the keyword arguments for Plots.jl.

PowerDynSolve.GridSolutionError — Type.

Error to be thrown if something goes wrong during when solving a power grid model.

PowerDynSolve.tslabel — Method.

Create the standard variable labels for power grid plots.

PowerDynSolve.tstransform — Method.

Transform the array output from DifferentialEquations.jl correctly to be used in Plots.jl's recipes.

DiffEqBase.solve — Function.

function solve(p::GridProblem; kwargs...)

Solve a PowerDynSolve.GridProblem, using DifferentialEquations.jl in the back. The correct solvers are automatically chosen.

kwargs are the keyword arguments that are simply passed to DifferentialEquations.jl.

function solve(g::GridDynamics, x0, timespan)

Solve a power grid g (of type PowerDynSolve.GridDynamics) starting at x0 for a timespan, using DifferentialEquations.jl in the back. The correct solvers are automatically chosen.