OpenTrick.jl Documentation

OpenTrick.jl Documentation

OpenTrick.opentrick — Function.

opentrick(openfn[, args... [; <keyword arguments>]])

Call openfn with (handlefn, args... ,kwargs ...) as arguments, return an IOWrapper instance. (NB:handlefn is provided by opentrick.)

Arguments

openfn::Function function actually called to obtain a IO instance. openfn must take a Function(::IO) instance as its first argument

args optional arguments that will be passed to openfn

kwargs optional keyword arguments that will be passed to openfn

Examples

julia> using OpenTrick

julia> filename = tempname();

julia> io = opentrick(open, filename, "w+");

julia> write(io, "hello world!")
12

julia> seek(io, 0);

julia> readline(io)
"hello world!"

OpenTrick.rawio — Function.

rawio(io)

Return the actual io instance

OpenTrick.blockingtask — Function.

blockingtask(io)

Return the task blocking which prevents the handlefn passed to openfn from returning

OpenTrick.unsafe_clear — Function.

unsafe_clear()

Unblock all blocking tasks. All ios returned by opentrick will be closed as a consequence.