Bitcoin.jl Documentation

Bitcoin.jl Documentation

A Bitcoin library for Julia

Functions

Address

Bitcoin.address — Function.

adress(P::ECC.S256Point, compressed::Bool, testnet::Bool) -> String

Returns the Base58 Bitcoin address given an S256Point Compressed is set to true if not provided. Testnet is set to false by default.

Bitcoin.wif — Function.

wif(pk::PrivateKey, compressed::Bool=true, testnet::Bool=false) -> String

Returns a PrivateKey in Wallet Import Format Compressed is set to true if not provided. Testnet is set to false by default.

Transaction

Bitcoin.txparse — Function.

txparse(s::Base.GenericIOBuffer, testnet::Bool=false) -> Tx

Returns a Tx object given a byte stream

Bitcoin.txserialize — Function.

txserialize(tx::Tx) -> Array{UInt8,1}

Returns the byte serialization of the transaction

Bitcoin.txfetch — Function.

txfetch(tx_id::String) -> Tx

Returns the bitcoin transaction given its ID as an hexadecimal string.

Bitcoin.txid — Function.

txid(tx::Tx) -> String

Returns an hexadecimal string of the transaction hash

Bitcoin.txfee — Function.

txfee(tx::Tx) -> Integer

Returns the fee of this transaction in satoshi

Bitcoin.txsighash — Function.

txsighash(tx::Tx, input_index::Integer) -> Integer

Returns the integer representation of the hash that needs to get signed for index input_index

Script

Bitcoin.scriptevaluate — Function.

scriptevaluate(s::Script, z::Integer) -> Bool

Evaluate if Script is valid given the transaction signature hash

Buy me a cup of coffee

Donate Bitcoin Donate Litecoin

Index

Bitcoin.address

Bitcoin.scriptevaluate

Bitcoin.txfee

Bitcoin.txfetch

Bitcoin.txid

Bitcoin.txparse

Bitcoin.txserialize

Bitcoin.txsighash

Bitcoin.wif