Home

GPUifyLoops.jl

GPUifyLoops tries to solve the problem of code-duplication that can occur when writing performant kernels that target multiple devices.

API

GPUifyLoops.@loop — Macro.

@loop for i in (A; B)
    # body
end

Take a for i in (A; B) expression and on the CPU lowers it to:

for i in A
    # body
end

and on the GPU:

for i in B
    if !(i in A)
        continue
    end
    # body
end

GPUifyLoops.@setup — Macro.

@setup Dev

Setups some hidden state within the function that allows the other macros to properly work.

function kernel(::Val{Dev}, A) where Dev
    @setup Dev
    # ...
end

kernel(A::Array) = kernel(Val(:CPU), A)
kernel(A::CuArray) = @cuda kernel(Val(:GPU), A)

GPUifyLoops.@synchronize — Macro.

@syncronize

Calls sync_threads() on the GPU and nothing on the CPU.

Examples

Simple

using GPUifyLoops

kernel(A::Array) = kernel(Val(:CPU), A)
function kernel(::Val{Dev}, A) where Dev
    @setup Dev

    @loop for i in (1:size(A,1);
                    threadIdx().x)
        A[i] = 2*A[i]
    end
    @synchronize
end

data = Array{Float32}(undef, 1024)
kernel(data)

@static if Base.find_package("CuArrays") !== nothing
    using CuArrays
    using CUDAnative

    @eval function kernel(A::CuArray)
        @cuda threads=length(A) kernel(Val(:GPU), A)
    end

    data = CuArray{Float32}(undef, 1024)
    kernel(data)
end