Docstrings

LineSearches.@pack_BackTracking

LineSearches.@pack_BackTracking!

LineSearches.@pack_HagerZhang

LineSearches.@pack_HagerZhang!

LineSearches.@pack_InitialHagerZhang

LineSearches.@pack_InitialHagerZhang!

LineSearches.@pack_InitialPrevious

LineSearches.@pack_InitialPrevious!

LineSearches.@pack_InitialQuadratic

LineSearches.@pack_InitialQuadratic!

LineSearches.@pack_InitialStatic

LineSearches.@pack_InitialStatic!

LineSearches.@pack_MoreThuente

LineSearches.@pack_MoreThuente!

LineSearches.@pack_StrongWolfe

LineSearches.@pack_StrongWolfe!

LineSearches.@unpack_BackTracking

LineSearches.@unpack_HagerZhang

LineSearches.@unpack_InitialHagerZhang

LineSearches.@unpack_InitialPrevious

LineSearches.@unpack_InitialQuadratic

LineSearches.@unpack_InitialStatic

LineSearches.@unpack_MoreThuente

LineSearches.@unpack_StrongWolfe

LineSearches.ALPHA

LineSearches.BARRIERCOEF

LineSearches.BETA

LineSearches.BISECT

LineSearches.BRACKET

LineSearches.BackTracking — Type.

BackTracking specifies a backtracking line-search that uses a quadratic or cubic interpolant to determine the reduction in step-size. E.g., if f(α) > f(0) + c₁ α f'(0), then the quadratic interpolant of f(0), f'(0), f(α) has a minimiser α' in the open interval (0, α). More strongly, there exists a factor ρ = ρ(c₁) such that α' ≦ ρ α.

This is a modification of the algorithm described in Nocedal Wright (2nd ed), Sec. 3.5.

source

LineSearches.DEFAULTDELTA

LineSearches.DEFAULTSIGMA

LineSearches.FINAL

LineSearches.GRADIENT

LineSearches.HagerZhang — Type.

Conjugate gradient line search implementation from: W. W. Hager and H. Zhang (2006) Algorithm 851: CG_DESCENT, a conjugate gradient method with guaranteed descent. ACM Transactions on Mathematical Software 32: 113–137.

source

LineSearches.ITER

LineSearches.InitialConstantChange — Type.

Constant first-order change approximation to determine initial step length.

** This requires that the optimization algorithm stores dphi0 from the previous iteration ** (dphi0previous = real(dot(∇f{k-1}, s_{k-1})), where s is the step direction.

This is meant for methods that do not produce well-scaled search directions, such as Gradient Descent and (variations of) Conjugate Gradient methods. See the discussion in Nocedal and Wright, 2nd ed, p. 59 on "Initial Step Length"

This procedure have several arguments, with the following defaults.

α0 = 1.0. The initial step size at the first iteration.

αmin = 1e-12. The minimum initial step size. (Default arbitrary).

αmax = 1.0. The maximum initial step size.

ρ = 0.25. Maximum decrease from previous iteration, αinit ≥ α_{k-1}. (Default arbitrary).

snap2one = (0.75, Inf). Set all values within this (closed) interval to 1.0. (Default arbitrary).

If αmax ≠ 1.0, then you should consider to ensure that snap2one[2] < αmax.

LineSearches.InitialHagerZhang — Type.

Initial step size algorithm from W. W. Hager and H. Zhang (2006) Algorithm 851: CG_DESCENT, a conjugate gradient method with guaranteed descent. ACM Transactions on Mathematical Software 32: 113–137.

If α0 is NaN, then procedure I0 is called at the first iteration, otherwise, we select according to procedure I1-2, with starting value α0.

source

LineSearches.InitialPrevious — Type.

Use previous step length as initial guess, within the bounds [alphamin, alphamax]

If state.alpha is NaN, then return fallback value is.alpha

source

LineSearches.InitialQuadratic — Type.

Quadratic interpolation for initial step length guess.

This is meant for methods that do not produce well-scaled search directions, such as Gradient Descent and (variations of) Conjugate Gradient methods. See the discussion around Nocedal and Wright, 2nd ed, (3.60).

This procedure have several arguments, with the following defaults.

α0 = 1.0. The initial step size at the first iteration.

αmin = 1e-12. The minimum initial step size. (Default arbitrary).

αmax = 1.0. The maximum initial step size.

ρ = 0.25. Maximum decrease from previous iteration, αinit ≥ α_{k-1}. (Default arbitrary).

snap2one = (0.75, Inf). Set all values within this (closed) interval to 1.0. (Default arbitrary).

If αmax ≠ 1.0, then you should consider to ensure that snap2one[2] < αmax.

source

LineSearches.InitialStatic — Type.

Provide static initial step length.

Keyword alpha corresponds to static step length, default is 1.0. If keyword scaled = true, then the initial step length is scaled with the l_2 norm of the step direction.

source

LineSearches.LINESEARCH

LineSearches.LineSearchException

LineSearches.LineSearches

LineSearches.MoreThuente — Type.

The line search implementation from: Moré, Jorge J., and David J. Thuente Line search algorithms with guaranteed sufficient decrease. ACM Transactions on Mathematical Software (TOMS) 20.3 (1994): 286-307.

source

LineSearches.PARAMETERS

LineSearches.SEARCHDIR

LineSearches.SECANT2

LineSearches.Static — Type.

Static: defines a static linesearch which returns the initial step length.

Static is intended for methods with well-scaled updates; i.e. Newton, on well-behaved problems.

LineSearches.StrongWolfe — Type.

StrongWolfe: This linesearch algorithm guarantees that the step length satisfies the (strong) Wolfe conditions. See Nocedal and Wright - Algorithms 3.5 and 3.6

This algorithm is mostly of theoretical interest, users should most likely use MoreThuente, HagerZhang or BackTracking.

Parameters: (and defaults)

c_1 = 1e-4: Armijo condition

c_2 = 0.9 : second (strong) Wolfe condition

ρ = 2.0 : bracket growth

source

LineSearches.UPDATE

LineSearches._hzI0

LineSearches._hzI12

LineSearches.bisect!

LineSearches.cstep

LineSearches.display_nextbit

LineSearches.eval

LineSearches.get_finite

LineSearches.include

LineSearches.interpolate

LineSearches.make_ϕ

LineSearches.make_ϕ_dϕ

LineSearches.make_ϕ_dϕ_ϕdϕ

LineSearches.make_ϕ_ϕdϕ

LineSearches.make_ϕdϕ

LineSearches.one64

LineSearches.satisfies_wolfe

LineSearches.secant

LineSearches.secant2!

LineSearches.update!

LineSearches.zoom