Docstrings

StochasticDiffEq.@cache

StochasticDiffEq.@iipnlcachefields

StochasticDiffEq.@iipnlsolve

StochasticDiffEq.@oopnlcachefields

StochasticDiffEq.@oopnlsolve

StochasticDiffEq.@tight_loop_macros

StochasticDiffEq.AbstractAnalyticalProblem

StochasticDiffEq.AbstractNLsolveCache

StochasticDiffEq.AbstractNLsolveSolver

StochasticDiffEq.AffineDiffEqOperator

StochasticDiffEq.AnalyticalProblem

StochasticDiffEq.AutoAlgSwitch

StochasticDiffEq.AutoSOSRA2

StochasticDiffEq.AutoSOSRI2

StochasticDiffEq.AutoSwitch

StochasticDiffEq.BVProblem

StochasticDiffEq.CallbackSet

StochasticDiffEq.CompiledFloats

StochasticDiffEq.ContinuousCallback

StochasticDiffEq.DAEFunction

StochasticDiffEq.DAEProblem

StochasticDiffEq.DAESolution

StochasticDiffEq.DDEFunction

StochasticDiffEq.DDEProblem

StochasticDiffEq.DEDataArray

StochasticDiffEq.DEDataMatrix

StochasticDiffEq.DEDataVector

StochasticDiffEq.DEFAULT_LINSOLVE

StochasticDiffEq.DESolution

StochasticDiffEq.DiffCache

StochasticDiffEq.DiffEqBase

StochasticDiffEq.DiffEqNLSolveTag

StochasticDiffEq.DiscreteCallback

StochasticDiffEq.DiscreteFunction

StochasticDiffEq.DiscreteProblem

StochasticDiffEq.DynamicalODEFunction

StochasticDiffEq.DynamicalODEProblem

StochasticDiffEq.EM

StochasticDiffEq.EMCache

StochasticDiffEq.EMConstantCache

StochasticDiffEq.EulerHeun

StochasticDiffEq.EulerHeunCache

StochasticDiffEq.EulerHeunConstantCache

StochasticDiffEq.FourStageSRICache

StochasticDiffEq.FourStageSRIConstantCache

StochasticDiffEq.IIF1M

StochasticDiffEq.IIF1MCache

StochasticDiffEq.IIF1MConstantCache

StochasticDiffEq.IIF1Mil

StochasticDiffEq.IIF1MilCache

StochasticDiffEq.IIF1MilConstantCache

StochasticDiffEq.IIF2M

StochasticDiffEq.IIF2MCache

StochasticDiffEq.IIF2MConstantCache

StochasticDiffEq.ISSEM

StochasticDiffEq.ISSEMCache

StochasticDiffEq.ISSEMConstantCache

StochasticDiffEq.ISSEulerHeun

StochasticDiffEq.ISSEulerHeunCache

StochasticDiffEq.ISSEulerHeunConstantCache

StochasticDiffEq.ImplicitEM

StochasticDiffEq.ImplicitEMCache

StochasticDiffEq.ImplicitEMConstantCache

StochasticDiffEq.ImplicitEulerHeun

StochasticDiffEq.ImplicitEulerHeunCache

StochasticDiffEq.ImplicitEulerHeunConstantCache

StochasticDiffEq.ImplicitRKMil

StochasticDiffEq.ImplicitRKMilCache

StochasticDiffEq.ImplicitRKMilConstantCache

StochasticDiffEq.LambaEM

StochasticDiffEq.LambaEMCache

StochasticDiffEq.LambaEMConstantCache

StochasticDiffEq.LambaEulerHeun

StochasticDiffEq.LambaEulerHeunCache

StochasticDiffEq.LambaEulerHeunConstantCache

StochasticDiffEq.LinSolveFactorize

StochasticDiffEq.LinearInterpolationData

StochasticDiffEq.MonteCarloProblem

StochasticDiffEq.MonteCarloSolution

StochasticDiffEq.MonteCarloSummary

StochasticDiffEq.MonteCarloTestSolution

StochasticDiffEq.NLAnderson

StochasticDiffEq.NLFunctional — Type.

(S::NLFunctional)(integrator) -> (z, η, iter, fail_convergence)

Perform functional iteration that is used by implicit methods, where z is the solution, η is used to measure the iteration error (see [HW96]), iter is the number of iteration, and fail_convergence reports whether the algorithm succeed. It solves

\[G(z) = dt⋅f(tmp + γ⋅z, p, t+c⋅h) z = G(z)\]

by iterating

\[zᵏ⁺¹ = G(zᵏ).\]

[HW96]
Ernst Hairer and Gerhard Wanner, "Solving Ordinary Differential

Equations II, Springer Series in Computational Mathematics. ISBN 978-3-642-05221-7. Section IV.8. doi:10.1007/978-3-642-05221-7

StochasticDiffEq.NLNewton — Type.

(S::NLNewton)(integrator) -> (z, η, iter, fail_convergence)

Perform numerically stable modified NLNewton iteration that is specialized for implicit methods (see [HS96] and [HW96]), where z is the solution, η is used to measure the iteration error (see [HW96]), iter is the number of iteration, and fail_convergence reports whether the algorithm succeed. It solves

\[G(z) = dt⋅f(tmp + γ⋅z, p, t+c⋅h) - z = 0⃗\]

by iterating

\[W Δᵏ = dt⋅f(tmp + γ⋅zᵏ, p, t+c⋅h) - zᵏ zᵏ⁺¹ = zᵏ + Δᵏ\]

where W=M-dt⋅γJ, M is the mass matrix, dt is the step size, γ is a constant, and J is the Jacobian matrix.

[HS96]
M.E.Hoseaa and L.F.Shampine, "Analysis and implementation of TR-BDF2",

Applied Numerical Mathematics, Volume 20, Issues 1–2, February 1996, Pages 21-37. doi:10.1016/0168-9274(95)00115-8

[HW96]
Ernst Hairer and Gerhard Wanner, "Solving Ordinary Differential

Equations II, Springer Series in Computational Mathematics. ISBN 978-3-642-05221-7. Section IV.8. doi:10.1007/978-3-642-05221-7

StochasticDiffEq.NLSOLVEJL_SETUP

StochasticDiffEq.NLSolverCache

StochasticDiffEq.NoiseProblem

StochasticDiffEq.ODEFunction

StochasticDiffEq.ODEProblem

StochasticDiffEq.ODESolution

StochasticDiffEq.PCEuler — Type.

PCEuler(ggprime; theta=1/2, eta=1/2)

Predictor Corrector Euler

Arguments

ggprime::Function: For scalar problems, ggprime $= b\partial_x(b)$ For multi-dimensional problems bbprime_k $= \sum_{j=1...M, i=1...D} b^(j)_i \partial_i b^(j)_k$ where $b^(j)$ correspond to the noise vector due to the j'th noise channel. If problem is in place - a in place ggprime should be supplied - and vice versa for not in place speicification of problem.

theta::Real: Degree of implicitness in the drift term. Set to 0.5 by default.

eta::Real: Degree of implicitness in the diffusion term. Set to 0.5 by default.

Reference: Stochastics and Dynamics, Vol. 8, No. 3 (2008) 561–581 Note that the original paper has a typo in the definition of ggprime...

StochasticDiffEq.PCEulerCache

StochasticDiffEq.PCEulerConstantCache

StochasticDiffEq.RHS_IIF1

StochasticDiffEq.RHS_IIF1M_Scalar

StochasticDiffEq.RHS_IIF2

StochasticDiffEq.RHS_IIF2M_Scalar

StochasticDiffEq.RKMil

StochasticDiffEq.RKMilCache

StochasticDiffEq.RKMilCommute

StochasticDiffEq.RKMilCommuteCache

StochasticDiffEq.RKMilCommuteConstantCache

StochasticDiffEq.RKMilConstantCache

StochasticDiffEq.RODECompositeSolution

StochasticDiffEq.RODEFunction

StochasticDiffEq.RODEProblem

StochasticDiffEq.RODESolution

StochasticDiffEq.RandomEM

StochasticDiffEq.RandomEMCache

StochasticDiffEq.RandomEMConstantCache

StochasticDiffEq.RosslerSRA — Type.

RosslerSRA

Holds the Butcher tableaus for a Rosser SRA method.

StochasticDiffEq.RosslerSRI — Type.

RosslerSRI

Holds the Butcher tableaus for a Rosser SRI method.

StochasticDiffEq.SDEFunction

StochasticDiffEq.SDEIntegrator

StochasticDiffEq.SDEOptions

StochasticDiffEq.SDEProblem

StochasticDiffEq.SKenCarp

StochasticDiffEq.SKenCarpCache

StochasticDiffEq.SKenCarpConstantCache

StochasticDiffEq.SKenCarpTableau

StochasticDiffEq.SOSRA

StochasticDiffEq.SOSRA2

StochasticDiffEq.SOSRA2ConstantCache

StochasticDiffEq.SOSRAConstantCache

StochasticDiffEq.SOSRI

StochasticDiffEq.SOSRI2

StochasticDiffEq.SOSRI2ConstantCache

StochasticDiffEq.SOSRIConstantCache

StochasticDiffEq.SRA

StochasticDiffEq.SRA1

StochasticDiffEq.SRA1Cache

StochasticDiffEq.SRA1ConstantCache

StochasticDiffEq.SRA2

StochasticDiffEq.SRA2Cache

StochasticDiffEq.SRA2ConstantCache

StochasticDiffEq.SRA3

StochasticDiffEq.SRA3ConstantCache

StochasticDiffEq.SRACache

StochasticDiffEq.SRAConstantCache

StochasticDiffEq.SRI

StochasticDiffEq.SRICache

StochasticDiffEq.SRIConstantCache

StochasticDiffEq.SRIW1

StochasticDiffEq.SRIW1Cache

StochasticDiffEq.SRIW1ConstantCache

StochasticDiffEq.SRIW2

StochasticDiffEq.SRIW2ConstantCache

StochasticDiffEq.SecondOrderODEProblem

StochasticDiffEq.SplitEM

StochasticDiffEq.SplitEMCache

StochasticDiffEq.SplitEMConstantCache

StochasticDiffEq.SplitFunction

StochasticDiffEq.SplitODEProblem

StochasticDiffEq.SplitSDEFunction

StochasticDiffEq.SplitSDEProblem

StochasticDiffEq.SteadyStateProblem

StochasticDiffEq.SteadyStateSolution

StochasticDiffEq.StochasticCompositeAlgorithm

StochasticDiffEq.StochasticCompositeCache

StochasticDiffEq.StochasticDiffEq

StochasticDiffEq.StochasticDiffEqAdaptiveAlgorithm

StochasticDiffEq.StochasticDiffEqAlgorithm

StochasticDiffEq.StochasticDiffEqCache

StochasticDiffEq.StochasticDiffEqCompositeAlgorithm

StochasticDiffEq.StochasticDiffEqConstantCache

StochasticDiffEq.StochasticDiffEqMutableCache

StochasticDiffEq.StochasticDiffEqNewtonAdaptiveAlgorithm

StochasticDiffEq.StochasticDiffEqNewtonAlgorithm

StochasticDiffEq.StochasticDiffEqRODEAdaptiveAlgorithm

StochasticDiffEq.StochasticDiffEqRODEAlgorithm

StochasticDiffEq.StochasticDiffEqRODECompositeAlgorithm

StochasticDiffEq.ThreeStageSRACache

StochasticDiffEq.ThreeStageSRAConstantCache

StochasticDiffEq.TimeChoiceIterator

StochasticDiffEq.TwoPointBVProblem

StochasticDiffEq.WOperator — Type.

WOperator(mass_matrix,gamma,J)

A linear operator that represents the W matrix of an ODEProblem, defined as

\[W = MM - \gamma J\]

where MM is the mass matrix (a regular AbstractMatrix or a UniformScaling), γ is a real number proportional to the time step, and J is the Jacobian operator (must be a AbstractDiffEqLinearOperator). A WOperator can also be constructed using a *DEFunction directly as

WOperator(f,gamma)

f needs to have a jacobian and jac_prototype, but the prototype does not need to be a diffeq operator –- it will automatically be converted to one.

WOperator supports lazy * and mul! operations, the latter utilizing an internal cache (can be specified in the constructor; default to regular Vector). It supports all of AbstractDiffEqLinearOperator's interface.

StochasticDiffEq._reshape

StochasticDiffEq._vec

StochasticDiffEq.add_saveat!

StochasticDiffEq.add_tstop!

StochasticDiffEq.addat!

StochasticDiffEq.addat_noise!

StochasticDiffEq.addat_non_user_cache!

StochasticDiffEq.addsteps!

StochasticDiffEq.alg_autodiff

StochasticDiffEq.alg_cache

StochasticDiffEq.alg_compatible

StochasticDiffEq.alg_interpretation

StochasticDiffEq.alg_mass_matrix_compatible

StochasticDiffEq.alg_needs_extra_process

StochasticDiffEq.alg_order

StochasticDiffEq.alg_stability_size

StochasticDiffEq.apply_step!

StochasticDiffEq.auto_dt_reset!

StochasticDiffEq.beta1_default

StochasticDiffEq.beta2_default

StochasticDiffEq.build_jac_config

StochasticDiffEq.calc_J — Function.

calc_J!(integrator,cache,is_compos)

Interface for calculating the jacobian.

For constant caches, a new jacobian object is returned whereas for mutable caches cache.J is updated. In both cases, if integrator.f has a custom jacobian update function, then it will be called for the update. Otherwise, either ForwardDiff or finite difference will be used depending on the jac_config of the cache.

StochasticDiffEq.calc_J!

StochasticDiffEq.calc_W!

StochasticDiffEq.calc_dt_propose!

StochasticDiffEq.calc_tderivative

StochasticDiffEq.calc_tderivative!

StochasticDiffEq.calculate_residuals — Function.

calculate_residuals(ũ, u₀, u₁, α, ρ, scalarnorm)

Return element-wise residuals

\[\frac{ũ}{α+\max{scalarnorm(u₀),scalarnorm(u₁)}*ρ}.\]

calculate_residuals(E₁, E₂, u₀, u₁, α, ρ, δ, scalarnorm)

Return element-wise residuals

\[\frac{δ E₁ + E₂}{α+\max{scalarnorm(u₀),scalarnorm(u₁)}*ρ}.\]

StochasticDiffEq.calculate_residuals! — Function.

calculate_residuals!(out, ũ, u₀, u₁, α, ρ, scalarnorm)

Same as calculate_residuals but save result in out.

calculate_residuals!(out, E₁, E₂, u₀, u₁, α, ρ, δ, scalarnorm)

Same as calculate_residuals but save result in out.

StochasticDiffEq.change_t_via_interpolation!

StochasticDiffEq.checkSRAOrder — Function.

checkSRAOrder(RosslerSRI)

Determines whether the order conditions are met via the tableaus of the SRA method.

StochasticDiffEq.checkSRIOrder — Function.

checkSRIOrder(RosslerSRI)

Determines whether the order conditions are met via the tableaus of the SRI method.

StochasticDiffEq.check_error

StochasticDiffEq.check_keywords

StochasticDiffEq.choose_algorithm!

StochasticDiffEq.constructExplicitSKenCarp — Function.

constructSOSRA2()

Constructs the taleau type for the SOSRA method.

StochasticDiffEq.constructSKenCarp — Function.

constructSOSRA2()

Constructs the taleau type for the SOSRA method.

StochasticDiffEq.constructSOSRA — Function.

constructSOSRA()

Constructs the taleau type for the SOSRA method.

StochasticDiffEq.constructSOSRA2 — Function.

constructSOSRA2()

Constructs the taleau type for the SOSRA method.

StochasticDiffEq.constructSRA1 — Function.

constructSRA1()

Constructs the taleau type for the SRA1 method.

StochasticDiffEq.constructSRA2 — Function.

constructSRA2()

Constructs the taleau type for the SRA2 method.

StochasticDiffEq.constructSRA3 — Function.

constructSRA3()

Constructs the taleau type for the SRA3 method.

StochasticDiffEq.constructSRIOpt1 — Function.

constructSRIOpt1()

Opti6-12-11-10-01-47

StochasticDiffEq.constructSRIOpt2 — Function.

constructSRIOpt2()

Opti6-12-11-10-01-47

StochasticDiffEq.constructSRIW1 — Function.

constructSRIW1()

Constructs the tableau type for the SRIW1 method.

StochasticDiffEq.constructSRIW2 — Function.

constructSRIW2()

Constructs the tableau type for the SRIW1 method.

StochasticDiffEq.current_extrapolant

StochasticDiffEq.current_extrapolant!

StochasticDiffEq.current_interpolant

StochasticDiffEq.current_interpolant!

StochasticDiffEq.deleteat!

StochasticDiffEq.deleteat_noise!

StochasticDiffEq.deleteat_non_user_cache!

StochasticDiffEq.derivative

StochasticDiffEq.derivative!

StochasticDiffEq.determine_chunksize

StochasticDiffEq.du_cache

StochasticDiffEq.eval

StochasticDiffEq.fill_new_noise_caches!

StochasticDiffEq.fix_dtnew_at_bounds!

StochasticDiffEq.full_cache

StochasticDiffEq.generate_tildes

StochasticDiffEq.get_chunksize

StochasticDiffEq.get_current_alg_autodiff

StochasticDiffEq.get_current_alg_order

StochasticDiffEq.get_dt

StochasticDiffEq.get_du

StochasticDiffEq.get_du!

StochasticDiffEq.get_proposed_dt

StochasticDiffEq.get_tmp_cache

StochasticDiffEq.handle_callback_modifiers!

StochasticDiffEq.handle_callbacks!

StochasticDiffEq.handle_dt!

StochasticDiffEq.handle_tstop!

StochasticDiffEq.has_exp

StochasticDiffEq.has_expmv

StochasticDiffEq.has_expmv!

StochasticDiffEq.has_ldiv

StochasticDiffEq.has_ldiv!

StochasticDiffEq.has_mul

StochasticDiffEq.has_mul!

StochasticDiffEq.include

StochasticDiffEq.init

StochasticDiffEq.initialize!

StochasticDiffEq.initialize_callbacks!

StochasticDiffEq.intervals

StochasticDiffEq.is_composite

StochasticDiffEq.is_constant

StochasticDiffEq.is_split_step

StochasticDiffEq.is_stiff

StochasticDiffEq.isadaptive

StochasticDiffEq.isdiscrete

StochasticDiffEq.isdtchangeable

StochasticDiffEq.isinplace

StochasticDiffEq.jac_iter

StochasticDiffEq.jacobian

StochasticDiffEq.jacobian!

StochasticDiffEq.loopfooter!

StochasticDiffEq.loopheader!

StochasticDiffEq.modify_dt_for_tstops!

StochasticDiffEq.modify_dtnew_for_tstops!

StochasticDiffEq.perform_step!

StochasticDiffEq.qmax_default

StochasticDiffEq.qmin_default

StochasticDiffEq.rand_cache

StochasticDiffEq.ratenoise_cache

StochasticDiffEq.reeval_internals_due_to_modification!

StochasticDiffEq.reinit!

StochasticDiffEq.remake

StochasticDiffEq.reset_alg_dependent_opts! — Function.

If no user default, then this will change the default to the defaults for the second algorithm.

StochasticDiffEq.resize!

StochasticDiffEq.resize_noise!

StochasticDiffEq.resize_non_user_cache!

StochasticDiffEq.savevalues!

StochasticDiffEq.sde_determine_initdt

StochasticDiffEq.sde_extrapolant

StochasticDiffEq.sde_extrapolant!

StochasticDiffEq.sde_interpolant

StochasticDiffEq.sde_interpolant!

StochasticDiffEq.sde_interpolation — Function.

sde_interpolation(tvals,ts,timeseries,ks)

Get the value at tvals where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

sde_interpolation(tval::Number,ts,timeseries,ks)

Get the value at tval where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

StochasticDiffEq.sde_interpolation!

StochasticDiffEq.set_abstol!

StochasticDiffEq.set_gamma!

StochasticDiffEq.set_proposed_dt!

StochasticDiffEq.set_reltol!

StochasticDiffEq.set_t!

StochasticDiffEq.set_u!

StochasticDiffEq.solution_endpoint_match_cur_integrator!

StochasticDiffEq.solve

StochasticDiffEq.solve!

StochasticDiffEq.step!

StochasticDiffEq.terminate!

StochasticDiffEq.transfer_cache!

StochasticDiffEq.tstop_saveat_disc_handling

StochasticDiffEq.tuples

StochasticDiffEq.u_cache

StochasticDiffEq.u_modified!

StochasticDiffEq.unwrap_alg

StochasticDiffEq.update_coefficients

StochasticDiffEq.update_coefficients!

StochasticDiffEq.update_noise!

StochasticDiffEq.user_cache

StochasticDiffEq.warn_compat