Docstrings

OrdinaryDiffEq.@bs5pre0

OrdinaryDiffEq.@bs5pre1

OrdinaryDiffEq.@bs5unpack — Macro.

Coefficients taken from RKSuite

OrdinaryDiffEq.@cache

OrdinaryDiffEq.@dp5pre0 — Macro.

Hairer Norsett Wanner Solving Ordinary Differential Euations I - Nonstiff Problems Page 192

OrdinaryDiffEq.@dp5pre1

OrdinaryDiffEq.@dp5pre2

OrdinaryDiffEq.@dp5pre3

OrdinaryDiffEq.@dp5pre4

OrdinaryDiffEq.@dprkn6pre0

OrdinaryDiffEq.@dprkn6unpack — Macro.

OrdinaryDiffEq.@iipnlsolve

OrdinaryDiffEq.@oopnlsolve

OrdinaryDiffEq.@owrenzen3pre0

OrdinaryDiffEq.@owrenzen3pre1

OrdinaryDiffEq.@owrenzen3pre2

OrdinaryDiffEq.@owrenzen3pre3

OrdinaryDiffEq.@owrenzen3unpack — Macro.

OrdinaryDiffEq.@owrenzen4pre0

OrdinaryDiffEq.@owrenzen4pre1

OrdinaryDiffEq.@owrenzen4pre2

OrdinaryDiffEq.@owrenzen4pre3

OrdinaryDiffEq.@owrenzen4pre4

OrdinaryDiffEq.@owrenzen4unpack — Macro.

OrdinaryDiffEq.@owrenzen5pre0

OrdinaryDiffEq.@owrenzen5pre1

OrdinaryDiffEq.@owrenzen5pre2

OrdinaryDiffEq.@owrenzen5pre3

OrdinaryDiffEq.@owrenzen5pre4

OrdinaryDiffEq.@owrenzen5pre5

OrdinaryDiffEq.@owrenzen5unpack — Macro.

OrdinaryDiffEq.@rosenbrock2332pre0

OrdinaryDiffEq.@rosenbrock2332pre1

OrdinaryDiffEq.@rosenbrock2332unpack — Macro.

From MATLAB ODE Suite by Shampine

OrdinaryDiffEq.@ssprkpre0 — Macro.

Second order strong stability preserving (SSP) interpolant.

Ketcheson, Lóczi, Jangabylova, Kusmanov: Dense output for SSP RK methods (2017).

OrdinaryDiffEq.@ssprkpre1

OrdinaryDiffEq.@ssprkpre2

OrdinaryDiffEq.@swap!

OrdinaryDiffEq.@tight_loop_macros

OrdinaryDiffEq.@tsit5pre0

OrdinaryDiffEq.@tsit5pre1

OrdinaryDiffEq.@tsit5pre2

OrdinaryDiffEq.@tsit5pre3

OrdinaryDiffEq.@tsit5pre4

OrdinaryDiffEq.@tsit5unpack — Macro.

Runge–Kutta pairs of order 5(4) satisfying only the first column simplifying assumption

Ch. Tsitouras

OrdinaryDiffEq.@vern6pre0

OrdinaryDiffEq.@vern6pre1

OrdinaryDiffEq.@vern6unpack — Macro.

OrdinaryDiffEq.@vern7pre0

OrdinaryDiffEq.@vern7pre1

OrdinaryDiffEq.@vern7unpack — Macro.

OrdinaryDiffEq.@vern8pre0

OrdinaryDiffEq.@vern8pre1

OrdinaryDiffEq.@vern8unpack — Macro.

OrdinaryDiffEq.@vern9pre0

OrdinaryDiffEq.@vern9pre1

OrdinaryDiffEq.@vern9pre2

OrdinaryDiffEq.@vern9pre3

OrdinaryDiffEq.@vern9pre4

OrdinaryDiffEq.@vern9unpack — Macro.

OrdinaryDiffEq.AB3

OrdinaryDiffEq.AB3Cache

OrdinaryDiffEq.AB3ConstantCache

OrdinaryDiffEq.AB4

OrdinaryDiffEq.AB4Cache

OrdinaryDiffEq.AB4ConstantCache

OrdinaryDiffEq.AB5

OrdinaryDiffEq.AB5Cache

OrdinaryDiffEq.AB5ConstantCache

OrdinaryDiffEq.ABDF2

OrdinaryDiffEq.ABDF2Cache

OrdinaryDiffEq.ABDF2ConstantCache

OrdinaryDiffEq.ABM32

OrdinaryDiffEq.ABM32Cache

OrdinaryDiffEq.ABM32ConstantCache

OrdinaryDiffEq.ABM43

OrdinaryDiffEq.ABM43Cache

OrdinaryDiffEq.ABM43ConstantCache

OrdinaryDiffEq.ABM54

OrdinaryDiffEq.ABM54Cache

OrdinaryDiffEq.ABM54ConstantCache

OrdinaryDiffEq.AN5

OrdinaryDiffEq.AN5Cache

OrdinaryDiffEq.AN5ConstantCache

OrdinaryDiffEq.AbstractAnalyticalProblem

OrdinaryDiffEq.AbstractNLSolver

OrdinaryDiffEq.AbstractNLSolverCache

OrdinaryDiffEq.AffineDiffEqOperator

OrdinaryDiffEq.AnalyticalProblem

OrdinaryDiffEq.Anas5

OrdinaryDiffEq.Anas5Cache

OrdinaryDiffEq.Anas5ConstantCache — Type.

An Optimized Runge-Kutta method for the solution of Orbital Problems by Z.A. Anastassi and T.E. Simos Journal of Computational and Applied Mathematics, Volume 175, Issue 1, 1 March 2005, Pages 1 to 9.

OrdinaryDiffEq.AutoAlgSwitch

OrdinaryDiffEq.AutoDP5

OrdinaryDiffEq.AutoSwitch

OrdinaryDiffEq.AutoTsit5

OrdinaryDiffEq.AutoVern6

OrdinaryDiffEq.AutoVern7

OrdinaryDiffEq.AutoVern8

OrdinaryDiffEq.AutoVern9

OrdinaryDiffEq.BS3

OrdinaryDiffEq.BS3Cache

OrdinaryDiffEq.BS3ConstantCache — Type.

constructBogakiShampine3()

Constructs the tableau object for the Bogakai-Shampine Order 2/3 method.

constructBogakiShampine3()

Constructs the tableau object for the Bogakai-Shampine Order 2/3 method.

OrdinaryDiffEq.BS5

OrdinaryDiffEq.BS5Cache

OrdinaryDiffEq.BS5ConstantCache — Type.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

OrdinaryDiffEq.BS5Interp — Function.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

Used in the lazy construction of the dense output

k9, k10, k11 are not computed until called in the dense routine

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

Used in the lazy construction of the dense output

k9, k10, k11 are not computed until called in the dense routine

OrdinaryDiffEq.BS5Interp_polyweights — Function.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

These coefficients are taken from RKSuite

Note that RKSuite has an error: r081 should be 0 and r011 should be 1. This is pretty easy to spot since the first order interpolation is linear from y₀.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

These coefficients are taken from RKSuite

Note that RKSuite has an error: r081 should be 0 and r011 should be 1. This is pretty easy to spot since the first order interpolation is linear from y₀.

OrdinaryDiffEq.BVProblem

OrdinaryDiffEq.CFRLDDRK64

OrdinaryDiffEq.CFRLDDRK64ConstantCache

OrdinaryDiffEq.CNAB2

OrdinaryDiffEq.CNAB2Cache

OrdinaryDiffEq.CNAB2ConstantCache

OrdinaryDiffEq.CNLF2

OrdinaryDiffEq.CNLF2Cache

OrdinaryDiffEq.CNLF2ConstantCache

OrdinaryDiffEq.CallbackSet

OrdinaryDiffEq.CalvoSanz4

OrdinaryDiffEq.CalvoSanz4ConstantCache

OrdinaryDiffEq.CandyRoz4

OrdinaryDiffEq.CandyRoz4ConstantCache

OrdinaryDiffEq.CarpenterKennedy2N54

OrdinaryDiffEq.CarpenterKennedy2N54ConstantCache

OrdinaryDiffEq.Cash4

OrdinaryDiffEq.Cash4Cache

OrdinaryDiffEq.Cash4ConstantCache

OrdinaryDiffEq.Cash4Tableau

OrdinaryDiffEq.CompiledFloats

OrdinaryDiffEq.CompositeAlgorithm

OrdinaryDiffEq.CompositeCache

OrdinaryDiffEq.CompositeInterpolationData

OrdinaryDiffEq.ContinuousCallback

OrdinaryDiffEq.DAEFunction

OrdinaryDiffEq.DAEProblem

OrdinaryDiffEq.DAESolution

OrdinaryDiffEq.DDEFunction

OrdinaryDiffEq.DDEProblem

OrdinaryDiffEq.DEDataArray

OrdinaryDiffEq.DEDataMatrix

OrdinaryDiffEq.DEDataVector

OrdinaryDiffEq.DEFAULT_LINSOLVE

OrdinaryDiffEq.DEOptions

OrdinaryDiffEq.DESolution

OrdinaryDiffEq.DGLDDRK73_C

OrdinaryDiffEq.DGLDDRK73_CConstantCache

OrdinaryDiffEq.DGLDDRK84_C

OrdinaryDiffEq.DGLDDRK84_CConstantCache

OrdinaryDiffEq.DGLDDRK84_F

OrdinaryDiffEq.DGLDDRK84_FConstantCache

OrdinaryDiffEq.DP5

OrdinaryDiffEq.DP5Cache

OrdinaryDiffEq.DP5ConstantCache

OrdinaryDiffEq.DP5Threaded

OrdinaryDiffEq.DP5ThreadedCache

OrdinaryDiffEq.DP5_dense_ds

OrdinaryDiffEq.DP8

OrdinaryDiffEq.DP8Cache

OrdinaryDiffEq.DP8ConstantCache

OrdinaryDiffEq.DP8Interp

OrdinaryDiffEq.DP8Interp_polyweights

OrdinaryDiffEq.DPRKN12

OrdinaryDiffEq.DPRKN12Cache

OrdinaryDiffEq.DPRKN12ConstantCache

OrdinaryDiffEq.DPRKN6

OrdinaryDiffEq.DPRKN6Cache

OrdinaryDiffEq.DPRKN6ConstantCache

OrdinaryDiffEq.DPRKN8

OrdinaryDiffEq.DPRKN8Cache

OrdinaryDiffEq.DPRKN8ConstantCache

OrdinaryDiffEq.DiffCache

OrdinaryDiffEq.DiffEqBase

OrdinaryDiffEq.DiffEqNLSolveTag

OrdinaryDiffEq.DiscreteCallback

OrdinaryDiffEq.DiscreteFunction

OrdinaryDiffEq.DiscreteProblem

OrdinaryDiffEq.DynamicalODEFunction

OrdinaryDiffEq.DynamicalODEProblem

OrdinaryDiffEq.EPIRK4s3A

OrdinaryDiffEq.EPIRK4s3ACache

OrdinaryDiffEq.EPIRK4s3AConstantCache

OrdinaryDiffEq.EPIRK4s3B

OrdinaryDiffEq.EPIRK4s3BCache

OrdinaryDiffEq.EPIRK4s3BConstantCache

OrdinaryDiffEq.EPIRK5P1

OrdinaryDiffEq.EPIRK5P1Cache

OrdinaryDiffEq.EPIRK5P1ConstantCache

OrdinaryDiffEq.EPIRK5P2

OrdinaryDiffEq.EPIRK5P2Cache

OrdinaryDiffEq.EPIRK5P2ConstantCache

OrdinaryDiffEq.EPIRK5s3

OrdinaryDiffEq.EPIRK5s3Cache

OrdinaryDiffEq.EPIRK5s3ConstantCache

OrdinaryDiffEq.ERKN4

OrdinaryDiffEq.ERKN4Cache

OrdinaryDiffEq.ERKN4ConstantCache

OrdinaryDiffEq.ERKN5

OrdinaryDiffEq.ERKN5Cache

OrdinaryDiffEq.ERKN5ConstantCache

OrdinaryDiffEq.ESDIRK4Tableau

OrdinaryDiffEq.ETD1

OrdinaryDiffEq.ETD2

OrdinaryDiffEq.ETD2Cache

OrdinaryDiffEq.ETD2ConstantCache

OrdinaryDiffEq.ETD2Fsal

OrdinaryDiffEq.ETDRK2

OrdinaryDiffEq.ETDRK2Cache

OrdinaryDiffEq.ETDRK2ConstantCache

OrdinaryDiffEq.ETDRK3

OrdinaryDiffEq.ETDRK3Cache

OrdinaryDiffEq.ETDRK3ConstantCache

OrdinaryDiffEq.ETDRK4

OrdinaryDiffEq.ETDRK4Cache

OrdinaryDiffEq.ETDRK4ConstantCache

OrdinaryDiffEq.EXPRB53s3

OrdinaryDiffEq.EXPRB53s3Cache

OrdinaryDiffEq.EXPRB53s3ConstantCache

OrdinaryDiffEq.Euler

OrdinaryDiffEq.EulerCache

OrdinaryDiffEq.EulerConstantCache

OrdinaryDiffEq.Exp4

OrdinaryDiffEq.Exp4Cache

OrdinaryDiffEq.Exp4ConstantCache

OrdinaryDiffEq.ExpRKCache

OrdinaryDiffEq.ExpRKConstantCache

OrdinaryDiffEq.ExplicitRK

OrdinaryDiffEq.ExplicitRKCache

OrdinaryDiffEq.ExplicitRKConstantCache

OrdinaryDiffEq.ExponentialAlgorithm

OrdinaryDiffEq.Exprb32

OrdinaryDiffEq.Exprb32Cache

OrdinaryDiffEq.Exprb32ConstantCache

OrdinaryDiffEq.Exprb43

OrdinaryDiffEq.Exprb43Cache

OrdinaryDiffEq.Exprb43ConstantCache

OrdinaryDiffEq.Feagin10

OrdinaryDiffEq.Feagin10Cache

OrdinaryDiffEq.Feagin10ConstantCache — Type.

constructFeagin10

constructFeagin10

OrdinaryDiffEq.Feagin12

OrdinaryDiffEq.Feagin12Cache

OrdinaryDiffEq.Feagin12ConstantCache — Type.

constructFeagin12

constructFeagin12

OrdinaryDiffEq.Feagin14

OrdinaryDiffEq.Feagin14Cache

OrdinaryDiffEq.Feagin14ConstantCache — Type.

constructFeagin14

constructFeagin14

OrdinaryDiffEq.FunctionMap

OrdinaryDiffEq.FunctionMapCache

OrdinaryDiffEq.FunctionMapConstantCache

OrdinaryDiffEq.FunctionMap_scale_by_time

OrdinaryDiffEq.GRK4A

OrdinaryDiffEq.GRK4AConstantCache

OrdinaryDiffEq.GRK4T

OrdinaryDiffEq.GRK4TConstantCache

OrdinaryDiffEq.GeneralRosenbrock

OrdinaryDiffEq.GenericIIF1

OrdinaryDiffEq.GenericIIF1Cache

OrdinaryDiffEq.GenericIIF1ConstantCache

OrdinaryDiffEq.GenericIIF2

OrdinaryDiffEq.GenericIIF2Cache

OrdinaryDiffEq.GenericIIF2ConstantCache

OrdinaryDiffEq.GenericImplicitEuler

OrdinaryDiffEq.GenericImplicitEulerCache

OrdinaryDiffEq.GenericImplicitEulerConstantCache

OrdinaryDiffEq.GenericTrapezoid

OrdinaryDiffEq.GenericTrapezoidCache

OrdinaryDiffEq.GenericTrapezoidConstantCache

OrdinaryDiffEq.HSLDDRK64

OrdinaryDiffEq.HSLDDRK64ConstantCache

OrdinaryDiffEq.Hairer4

OrdinaryDiffEq.Hairer42

OrdinaryDiffEq.Hairer42Tableau

OrdinaryDiffEq.Hairer4Cache

OrdinaryDiffEq.Hairer4ConstantCache

OrdinaryDiffEq.Hairer4Tableau

OrdinaryDiffEq.Heun

OrdinaryDiffEq.HeunCache

OrdinaryDiffEq.HeunConstantCache

OrdinaryDiffEq.HochOst4

OrdinaryDiffEq.HochOst4Cache

OrdinaryDiffEq.HochOst4ConstantCache

OrdinaryDiffEq.IMEXEuler

OrdinaryDiffEq.IRKN3

OrdinaryDiffEq.IRKN3Cache

OrdinaryDiffEq.IRKN3ConstantCache

OrdinaryDiffEq.IRKN4

OrdinaryDiffEq.IRKN4Cache

OrdinaryDiffEq.IRKN4ConstantCache

OrdinaryDiffEq.ImplicitEuler

OrdinaryDiffEq.ImplicitEulerCache

OrdinaryDiffEq.ImplicitEulerConstantCache

OrdinaryDiffEq.ImplicitMidpoint

OrdinaryDiffEq.ImplicitMidpointCache

OrdinaryDiffEq.ImplicitMidpointConstantCache

OrdinaryDiffEq.ImplicitRHS

OrdinaryDiffEq.ImplicitRHS_Scalar

OrdinaryDiffEq.InterpolationData

OrdinaryDiffEq.JVODE

OrdinaryDiffEq.JVODECache

OrdinaryDiffEq.JVODEConstantCache

OrdinaryDiffEq.JVODE_Adams

OrdinaryDiffEq.JVODE_BDF

OrdinaryDiffEq.KahanLi6

OrdinaryDiffEq.KahanLi6ConstantCache

OrdinaryDiffEq.KahanLi8

OrdinaryDiffEq.KahanLi8Cache

OrdinaryDiffEq.KahanLi8ConstantCache

OrdinaryDiffEq.KenCarp3

OrdinaryDiffEq.KenCarp3Cache

OrdinaryDiffEq.KenCarp3ConstantCache

OrdinaryDiffEq.KenCarp3Tableau

OrdinaryDiffEq.KenCarp4

OrdinaryDiffEq.KenCarp4Cache

OrdinaryDiffEq.KenCarp4ConstantCache

OrdinaryDiffEq.KenCarp4Tableau

OrdinaryDiffEq.KenCarp5

OrdinaryDiffEq.KenCarp5Cache

OrdinaryDiffEq.KenCarp5ConstantCache

OrdinaryDiffEq.KenCarp5Tableau

OrdinaryDiffEq.Kvaerno3

OrdinaryDiffEq.Kvaerno3Cache

OrdinaryDiffEq.Kvaerno3ConstantCache

OrdinaryDiffEq.Kvaerno3Tableau

OrdinaryDiffEq.Kvaerno4

OrdinaryDiffEq.Kvaerno4Cache

OrdinaryDiffEq.Kvaerno4ConstantCache

OrdinaryDiffEq.Kvaerno4Tableau

OrdinaryDiffEq.Kvaerno5

OrdinaryDiffEq.Kvaerno5Cache

OrdinaryDiffEq.Kvaerno5ConstantCache

OrdinaryDiffEq.Kvaerno5Tableau

OrdinaryDiffEq.LawsonEuler

OrdinaryDiffEq.LawsonEulerCache

OrdinaryDiffEq.LawsonEulerConstantCache

OrdinaryDiffEq.LinSolveFactorize

OrdinaryDiffEq.LinearExponential

OrdinaryDiffEq.LinearExponentialCache

OrdinaryDiffEq.LinearExponentialConstantCache

OrdinaryDiffEq.LowStorageRK2CCache

OrdinaryDiffEq.LowStorageRK2CConstantCache

OrdinaryDiffEq.LowStorageRK2NCache

OrdinaryDiffEq.LowStorageRK2NConstantCache

OrdinaryDiffEq.LowStorageRK3SCache

OrdinaryDiffEq.LowStorageRK3SConstantCache

OrdinaryDiffEq.MEBDF2

OrdinaryDiffEq.MEBDF2Cache

OrdinaryDiffEq.MEBDF2ConstantCache

OrdinaryDiffEq.MassMatrixAlgorithms

OrdinaryDiffEq.McAte2

OrdinaryDiffEq.McAte2ConstantCache

OrdinaryDiffEq.McAte3

OrdinaryDiffEq.McAte3ConstantCache

OrdinaryDiffEq.McAte4

OrdinaryDiffEq.McAte42

OrdinaryDiffEq.McAte42ConstantCache

OrdinaryDiffEq.McAte4ConstantCache

OrdinaryDiffEq.McAte5

OrdinaryDiffEq.McAte5ConstantCache

OrdinaryDiffEq.McAte8

OrdinaryDiffEq.McAte8Cache

OrdinaryDiffEq.McAte8ConstantCache

OrdinaryDiffEq.Midpoint

OrdinaryDiffEq.MidpointCache

OrdinaryDiffEq.MidpointConstantCache

OrdinaryDiffEq.MidpointSplitting

OrdinaryDiffEq.MidpointSplittingCache

OrdinaryDiffEq.MidpointSplittingConstantCache

OrdinaryDiffEq.MonteCarloProblem

OrdinaryDiffEq.MonteCarloSolution

OrdinaryDiffEq.MonteCarloSummary

OrdinaryDiffEq.MonteCarloTestSolution

OrdinaryDiffEq.MultistepAlgorithms

OrdinaryDiffEq.NDBLSRK124

OrdinaryDiffEq.NDBLSRK124ConstantCache

OrdinaryDiffEq.NDBLSRK134

OrdinaryDiffEq.NDBLSRK134ConstantCache

OrdinaryDiffEq.NDBLSRK144

OrdinaryDiffEq.NDBLSRK144ConstantCache

OrdinaryDiffEq.NLAnderson

OrdinaryDiffEq.NLAndersonCache

OrdinaryDiffEq.NLFunctional — Type.

(S::NLFunctional)(integrator) -> (z, η, iter, fail_convergence)

Perform functional iteration that is used by implicit methods, where z is the solution, η is used to measure the iteration error (see [HW96]), iter is the number of iteration, and fail_convergence reports whether the algorithm succeed. It solves

\[G(z) = dt⋅f(tmp + γ⋅z, p, t+c⋅h) z = G(z)\]

by iterating

\[zᵏ⁺¹ = G(zᵏ).\]

[HW96]
Ernst Hairer and Gerhard Wanner, "Solving Ordinary Differential

Equations II, Springer Series in Computational Mathematics. ISBN 978-3-642-05221-7. Section IV.8. doi:10.1007/978-3-642-05221-7

OrdinaryDiffEq.NLFunctionalCache

OrdinaryDiffEq.NLNewton — Type.

(S::NLNewton)(integrator) -> (z, η, iter, fail_convergence)

Perform numerically stable modified NLNewton iteration that is specialized for implicit methods (see [HS96] and [HW96]), where z is the solution, η is used to measure the iteration error (see [HW96]), iter is the number of iteration, and fail_convergence reports whether the algorithm succeed. It solves

\[G(z) = dt⋅f(tmp + γ⋅z, p, t+c⋅h) - z = 0⃗\]

by iterating

\[W Δᵏ = dt⋅f(tmp + γ⋅zᵏ, p, t+c⋅h) - zᵏ zᵏ⁺¹ = zᵏ + Δᵏ\]

where W=M-dt⋅γJ, M is the mass matrix, dt is the step size, γ is a constant, and J is the Jacobian matrix.

[HS96]
M.E.Hoseaa and L.F.Shampine, "Analysis and implementation of TR-BDF2",

Applied Numerical Mathematics, Volume 20, Issues 1–2, February 1996, Pages 21-37. doi:10.1016/0168-9274(95)00115-8

[HW96]
Ernst Hairer and Gerhard Wanner, "Solving Ordinary Differential

Equations II, Springer Series in Computational Mathematics. ISBN 978-3-642-05221-7. Section IV.8. doi:10.1007/978-3-642-05221-7

OrdinaryDiffEq.NLNewtonCache

OrdinaryDiffEq.NLSOLVEJL_SETUP

OrdinaryDiffEq.NoiseProblem

OrdinaryDiffEq.NorsettEuler

OrdinaryDiffEq.NorsettEulerCache

OrdinaryDiffEq.NorsettEulerConstantCache

OrdinaryDiffEq.Nystrom4

OrdinaryDiffEq.Nystrom4Cache

OrdinaryDiffEq.Nystrom4ConstantCache

OrdinaryDiffEq.Nystrom4VelocityIndependent

OrdinaryDiffEq.Nystrom4VelocityIndependentCache

OrdinaryDiffEq.Nystrom4VelocityIndependentConstantCache

OrdinaryDiffEq.Nystrom5VelocityIndependent

OrdinaryDiffEq.Nystrom5VelocityIndependentCache

OrdinaryDiffEq.Nystrom5VelocityIndependentConstantCache

OrdinaryDiffEq.NystromCCDefaultInitialization

OrdinaryDiffEq.NystromDefaultInitialization

OrdinaryDiffEq.ODEChunkCache

OrdinaryDiffEq.ODECompositeSolution

OrdinaryDiffEq.ODEEmptyCache

OrdinaryDiffEq.ODEFunction

OrdinaryDiffEq.ODEIntegrator — Type.

ODEIntegrator

Fundamental struct allowing interactively stepping through the numerical solving of a differential equation. The full documentation is hosted here: http://docs.juliadiffeq.org/latest/basics/integrator.html. This docstring describes basic functionality only!

Initialize using integrator = init(prob::ODEProblem, alg; kwargs...). The keyword args which are accepted are the same Common Solver Options used by solve.

For reference, relevant fields of the ODEIntegrator are:

t - time of the proposed step

u - value at the proposed step

opts - common solver options

alg - the algorithm associated with the solution

f - the function being solved

sol - the current state of the solution

tprev - the last timepoint

uprev - the value at the last timepoint

opts holds all of the common solver options, and can be mutated to change the solver characteristics. For example, to modify the absolute tolerance for the future timesteps, one can do:

integrator.opts.abstol = 1e-9

For more info see the linked documentation page.

OrdinaryDiffEq.ODEProblem

OrdinaryDiffEq.ODESolution

OrdinaryDiffEq.ODE_DEFAULT_TABLEAU — Constant.

ODEDEFAULTTABLEAU

Sets the default tableau for the ODE solver. Currently Dormand-Prince 4/5.

OrdinaryDiffEq.ORK256

OrdinaryDiffEq.ORK256ConstantCache

OrdinaryDiffEq.OrdinaryDiffEq

OrdinaryDiffEq.OrdinaryDiffEqAdamsVarOrderVarStepAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqAdaptiveAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqAdaptiveExponentialAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqAdaptiveImplicitAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqCache

OrdinaryDiffEq.OrdinaryDiffEqCompositeAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqConstantCache

OrdinaryDiffEq.OrdinaryDiffEqExponentialAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqImplicitAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqInterpolation

OrdinaryDiffEq.OrdinaryDiffEqMutableCache

OrdinaryDiffEq.OrdinaryDiffEqNewtonAdaptiveAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqNewtonAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqRosenbrockAdaptiveAlgorithm

OrdinaryDiffEq.OrdinaryDiffEqRosenbrockAlgorithm

OrdinaryDiffEq.OwrenZen3

OrdinaryDiffEq.OwrenZen3Cache

OrdinaryDiffEq.OwrenZen3ConstantCache

OrdinaryDiffEq.OwrenZen4

OrdinaryDiffEq.OwrenZen4Cache

OrdinaryDiffEq.OwrenZen4ConstantCache

OrdinaryDiffEq.OwrenZen5

OrdinaryDiffEq.OwrenZen5Cache

OrdinaryDiffEq.OwrenZen5ConstantCache

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S105

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S105ConstantCache

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S173

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S173ConstantCache

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S184

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S184ConstantCache

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S205

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S205ConstantCache

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S32

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S32ConstantCache

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S53

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S53ConstantCache

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S82

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S82ConstantCache

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S94

OrdinaryDiffEq.ParsaniKetchesonDeconinck3S94ConstantCache

OrdinaryDiffEq.PredictiveControllerAlgs

OrdinaryDiffEq.PseudoVerletLeapfrog

OrdinaryDiffEq.PseudoVerletLeapfrogConstantCache

OrdinaryDiffEq.QBDF

OrdinaryDiffEq.QBDF1

OrdinaryDiffEq.QBDF2

OrdinaryDiffEq.QNDF

OrdinaryDiffEq.QNDF1

OrdinaryDiffEq.QNDF1Cache

OrdinaryDiffEq.QNDF1ConstantCache

OrdinaryDiffEq.QNDF2

OrdinaryDiffEq.QNDF2Cache

OrdinaryDiffEq.QNDF2ConstantCache

OrdinaryDiffEq.QNDFCache

OrdinaryDiffEq.QNDFConstantCache

OrdinaryDiffEq.R!

OrdinaryDiffEq.RHS_IIF

OrdinaryDiffEq.RHS_IIF_Scalar

OrdinaryDiffEq.RK4

OrdinaryDiffEq.RK46NL

OrdinaryDiffEq.RK46NLCache

OrdinaryDiffEq.RK46NLConstantCache

OrdinaryDiffEq.RK4Cache

OrdinaryDiffEq.RK4ConstantCache

OrdinaryDiffEq.RKC

OrdinaryDiffEq.RKCCache

OrdinaryDiffEq.RKCConstantCache

OrdinaryDiffEq.ROCK2

OrdinaryDiffEq.ROCK2Cache

OrdinaryDiffEq.ROCK2ConstantCache

OrdinaryDiffEq.ROCK4

OrdinaryDiffEq.ROCK4Cache

OrdinaryDiffEq.ROCK4ConstantCache

OrdinaryDiffEq.RODEFunction

OrdinaryDiffEq.RODEProblem

OrdinaryDiffEq.RODESolution

OrdinaryDiffEq.ROS3P

OrdinaryDiffEq.ROS3PConstantCache

OrdinaryDiffEq.RadauIIA5

OrdinaryDiffEq.RadauIIA5Cache

OrdinaryDiffEq.RadauIIA5ConstantCache

OrdinaryDiffEq.RadauIIA5Tableau

OrdinaryDiffEq.Ralston

OrdinaryDiffEq.RalstonCache

OrdinaryDiffEq.RalstonConstantCache

OrdinaryDiffEq.RichardsonEuler

OrdinaryDiffEq.RichardsonEulerCache

OrdinaryDiffEq.RichardsonEulerConstantCache

OrdinaryDiffEq.Rodas3

OrdinaryDiffEq.Rodas3ConstantCache

OrdinaryDiffEq.Rodas4

OrdinaryDiffEq.Rodas42

OrdinaryDiffEq.Rodas42ConstantCache

OrdinaryDiffEq.Rodas4Cache

OrdinaryDiffEq.Rodas4ConstantCache

OrdinaryDiffEq.Rodas4P

OrdinaryDiffEq.Rodas4PConstantCache

OrdinaryDiffEq.Rodas5

OrdinaryDiffEq.Rodas5ConstantCache

OrdinaryDiffEq.RodasConstantCache

OrdinaryDiffEq.Ros4ConstantCache

OrdinaryDiffEq.Ros4LStab

OrdinaryDiffEq.Ros4LStabConstantCache

OrdinaryDiffEq.RosShamp4

OrdinaryDiffEq.RosShamp4ConstantCache

OrdinaryDiffEq.Rosenbrock23

OrdinaryDiffEq.Rosenbrock23Cache

OrdinaryDiffEq.Rosenbrock23ConstantCache

OrdinaryDiffEq.Rosenbrock32

OrdinaryDiffEq.Rosenbrock32Cache

OrdinaryDiffEq.Rosenbrock32ConstantCache

OrdinaryDiffEq.Rosenbrock33Cache

OrdinaryDiffEq.Rosenbrock33ConstantCache

OrdinaryDiffEq.Rosenbrock34Cache

OrdinaryDiffEq.Rosenbrock34ConstantCache

OrdinaryDiffEq.Rosenbrock4Cache

OrdinaryDiffEq.Rosenbrock4ConstantCache

OrdinaryDiffEq.Rosenbrock5Cache

OrdinaryDiffEq.Rosenbrock5ConstantCache

OrdinaryDiffEq.RosenbrockMutableCache

OrdinaryDiffEq.Ruth3

OrdinaryDiffEq.Ruth3ConstantCache

OrdinaryDiffEq.SBDF

OrdinaryDiffEq.SBDF2

OrdinaryDiffEq.SBDF3

OrdinaryDiffEq.SBDF4

OrdinaryDiffEq.SBDFCache

OrdinaryDiffEq.SBDFConstantCache

OrdinaryDiffEq.SDEFunction

OrdinaryDiffEq.SDEProblem

OrdinaryDiffEq.SDIRK2

OrdinaryDiffEq.SDIRK2Cache

OrdinaryDiffEq.SDIRK2ConstantCache

OrdinaryDiffEq.SDIRKMutableCache

OrdinaryDiffEq.SSPRK104

OrdinaryDiffEq.SSPRK104Cache

OrdinaryDiffEq.SSPRK104ConstantCache

OrdinaryDiffEq.SSPRK22

OrdinaryDiffEq.SSPRK22Cache

OrdinaryDiffEq.SSPRK22ConstantCache

OrdinaryDiffEq.SSPRK33

OrdinaryDiffEq.SSPRK33Cache

OrdinaryDiffEq.SSPRK33ConstantCache

OrdinaryDiffEq.SSPRK432

OrdinaryDiffEq.SSPRK432Cache

OrdinaryDiffEq.SSPRK432ConstantCache

OrdinaryDiffEq.SSPRK53

OrdinaryDiffEq.SSPRK53Cache

OrdinaryDiffEq.SSPRK53ConstantCache

OrdinaryDiffEq.SSPRK53_2N1

OrdinaryDiffEq.SSPRK53_2N1Cache

OrdinaryDiffEq.SSPRK53_2N1ConstantCache

OrdinaryDiffEq.SSPRK53_2N2

OrdinaryDiffEq.SSPRK53_2N2Cache

OrdinaryDiffEq.SSPRK53_2N2ConstantCache

OrdinaryDiffEq.SSPRK54

OrdinaryDiffEq.SSPRK54Cache

OrdinaryDiffEq.SSPRK54ConstantCache

OrdinaryDiffEq.SSPRK63

OrdinaryDiffEq.SSPRK63Cache

OrdinaryDiffEq.SSPRK63ConstantCache

OrdinaryDiffEq.SSPRK73

OrdinaryDiffEq.SSPRK73Cache

OrdinaryDiffEq.SSPRK73ConstantCache

OrdinaryDiffEq.SSPRK83

OrdinaryDiffEq.SSPRK83Cache

OrdinaryDiffEq.SSPRK83ConstantCache

OrdinaryDiffEq.SSPRK932

OrdinaryDiffEq.SSPRK932Cache

OrdinaryDiffEq.SSPRK932ConstantCache

OrdinaryDiffEq.SSPRKMSVS32

OrdinaryDiffEq.SSPRKMSVS32Cache

OrdinaryDiffEq.SSPRKMSVS32ConstantCache

OrdinaryDiffEq.SSPRKMSVS43

OrdinaryDiffEq.SSPRKMSVS43Cache

OrdinaryDiffEq.SSPRKMSVS43ConstantCache

OrdinaryDiffEq.SSPSDIRK2

OrdinaryDiffEq.SSPSDIRK2Cache

OrdinaryDiffEq.SSPSDIRK2ConstantCache

OrdinaryDiffEq.SecondOrderODEProblem

OrdinaryDiffEq.SofSpa10

OrdinaryDiffEq.SofSpa10Cache

OrdinaryDiffEq.SofSpa10ConstantCache

OrdinaryDiffEq.SplitAlgorithms

OrdinaryDiffEq.SplitEuler

OrdinaryDiffEq.SplitEulerCache

OrdinaryDiffEq.SplitEulerConstantCache

OrdinaryDiffEq.SplitFunction

OrdinaryDiffEq.SplitODEProblem

OrdinaryDiffEq.SplitSDEFunction

OrdinaryDiffEq.SplitSDEProblem

OrdinaryDiffEq.StandardControllerAlgs

OrdinaryDiffEq.SteadyStateProblem

OrdinaryDiffEq.SteadyStateSolution

OrdinaryDiffEq.Symplectic2Cache

OrdinaryDiffEq.Symplectic2ConstantCache

OrdinaryDiffEq.Symplectic3Cache

OrdinaryDiffEq.Symplectic3ConstantCache

OrdinaryDiffEq.Symplectic45Cache

OrdinaryDiffEq.Symplectic45ConstantCache

OrdinaryDiffEq.Symplectic4Cache

OrdinaryDiffEq.Symplectic4ConstantCache

OrdinaryDiffEq.Symplectic5Cache

OrdinaryDiffEq.Symplectic5ConstantCache

OrdinaryDiffEq.Symplectic62Cache

OrdinaryDiffEq.Symplectic62ConstantCache

OrdinaryDiffEq.Symplectic6Cache

OrdinaryDiffEq.Symplectic6ConstantCache

OrdinaryDiffEq.SymplecticEuler

OrdinaryDiffEq.SymplecticEulerCache

OrdinaryDiffEq.SymplecticEulerConstantCache

OrdinaryDiffEq.TRBDF2

OrdinaryDiffEq.TRBDF2Cache

OrdinaryDiffEq.TRBDF2ConstantCache

OrdinaryDiffEq.TRBDF2Tableau

OrdinaryDiffEq.TSLDDRK74

OrdinaryDiffEq.TSLDDRK74ConstantCache

OrdinaryDiffEq.TanYam7

OrdinaryDiffEq.TanYam7Cache

OrdinaryDiffEq.TanYam7ConstantCache — Type.

On the Optimization of Some Nine-Stage Seventh-order Runge-Kutta Method, by M. Tanaka, S. Muramatsu and S. Yamashita, Information Processing Society of Japan, Vol. 33, No. 12 (1992) pages 1512-1526.

On the Optimization of Some Nine-Stage Seventh-order Runge-Kutta Method, by M. Tanaka, S. Muramatsu and S. Yamashita, Information Processing Society of Japan, Vol. 33, No. 12 (1992) pages 1512-1526.

OrdinaryDiffEq.TimeChoiceIterator

OrdinaryDiffEq.Trapezoid

OrdinaryDiffEq.TrapezoidCache

OrdinaryDiffEq.TrapezoidConstantCache

OrdinaryDiffEq.Tsit5

OrdinaryDiffEq.Tsit5Cache

OrdinaryDiffEq.Tsit5ConstantCache

OrdinaryDiffEq.Tsit5Interp — Function.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

These are the coefficients of the expanded form of the polynomials from

Runge–Kutta pairs of order 5(4) satisfying only the first column simplifying assumption

Ch. Tsitouras

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

These are the coefficients of the expanded form of the polynomials from

Runge–Kutta pairs of order 5(4) satisfying only the first column simplifying assumption

Ch. Tsitouras

OrdinaryDiffEq.TsitPap8

OrdinaryDiffEq.TsitPap8Cache

OrdinaryDiffEq.TsitPap8ConstantCache — Type.

Cheap Error Estimation for Runge-Kutta methods, by Ch. Tsitouras and S.N. Papakostas, Siam Journal on Scientific Computing, Vol. 20, Issue 6, Nov 1999.

Cheap Error Estimation for Runge-Kutta methods, by Ch. Tsitouras and S.N. Papakostas, Siam Journal on Scientific Computing, Vol. 20, Issue 6, Nov 1999.

OrdinaryDiffEq.TwoPointBVProblem

OrdinaryDiffEq.U!

OrdinaryDiffEq.VCAB3

OrdinaryDiffEq.VCAB3Cache

OrdinaryDiffEq.VCAB3ConstantCache

OrdinaryDiffEq.VCAB4

OrdinaryDiffEq.VCAB4Cache

OrdinaryDiffEq.VCAB4ConstantCache

OrdinaryDiffEq.VCAB5

OrdinaryDiffEq.VCAB5Cache

OrdinaryDiffEq.VCAB5ConstantCache

OrdinaryDiffEq.VCABM

OrdinaryDiffEq.VCABM3

OrdinaryDiffEq.VCABM3Cache

OrdinaryDiffEq.VCABM3ConstantCache

OrdinaryDiffEq.VCABM4

OrdinaryDiffEq.VCABM4Cache

OrdinaryDiffEq.VCABM4ConstantCache

OrdinaryDiffEq.VCABM5

OrdinaryDiffEq.VCABM5Cache

OrdinaryDiffEq.VCABM5ConstantCache

OrdinaryDiffEq.VCABMCache

OrdinaryDiffEq.VCABMConstantCache

OrdinaryDiffEq.Veldd4

OrdinaryDiffEq.Veldd4ConstantCache

OrdinaryDiffEq.Velds4

OrdinaryDiffEq.Velds4ConstantCache

OrdinaryDiffEq.VelocityVerlet

OrdinaryDiffEq.VelocityVerletCache

OrdinaryDiffEq.VelocityVerletConstantCache

OrdinaryDiffEq.VerletLeapfrog

OrdinaryDiffEq.VerletLeapfrogConstantCache

OrdinaryDiffEq.Vern6

OrdinaryDiffEq.Vern6Cache

OrdinaryDiffEq.Vern6ConstantCache — Type.

From Verner's Website

From Verner's Website

OrdinaryDiffEq.Vern6Interp

OrdinaryDiffEq.Vern6Interp_polyweights — Function.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

OrdinaryDiffEq.Vern7

OrdinaryDiffEq.Vern7Cache

OrdinaryDiffEq.Vern7ConstantCache

OrdinaryDiffEq.Vern7Interp

OrdinaryDiffEq.Vern7Interp_polyweights

OrdinaryDiffEq.Vern8

OrdinaryDiffEq.Vern8Cache

OrdinaryDiffEq.Vern8ConstantCache

OrdinaryDiffEq.Vern8Interp

OrdinaryDiffEq.Vern8Interp_polyweights

OrdinaryDiffEq.Vern9

OrdinaryDiffEq.Vern9Cache

OrdinaryDiffEq.Vern9ConstantCache — Type.

From Verner's Webiste

From Verner's Webiste

OrdinaryDiffEq.Vern9Interp

OrdinaryDiffEq.Vern9Interp_polyweights

OrdinaryDiffEq.WOperator — Type.

WOperator(mass_matrix,gamma,J[;transform=false])

A linear operator that represents the W matrix of an ODEProblem, defined as

\[W = MM - \gamma J\]

or, if transform=true:

\[W = \frac{1}{\gamma}MM - J\]

where MM is the mass matrix (a regular AbstractMatrix or a UniformScaling), γ is a real number proportional to the time step, and J is the Jacobian operator (must be a AbstractDiffEqLinearOperator). A WOperator can also be constructed using a *DEFunction directly as

WOperator(f,gamma[;transform=false])

f needs to have a jacobian and jac_prototype, but the prototype does not need to be a diffeq operator –- it will automatically be converted to one.

WOperator supports lazy * and mul! operations, the latter utilizing an internal cache (can be specified in the constructor; default to regular Vector). It supports all of AbstractDiffEqLinearOperator's interface.

OrdinaryDiffEq.Yoshida6

OrdinaryDiffEq.Yoshida6ConstantCache

OrdinaryDiffEq._change_t_via_interpolation!

OrdinaryDiffEq._compute_nl

OrdinaryDiffEq._compute_nl!

OrdinaryDiffEq._ode_addsteps!

OrdinaryDiffEq._ode_interpolant — Function.

From MATLAB ODE Suite by Shampine

OrdinaryDiffEq._ode_interpolant!

OrdinaryDiffEq._phiv_timestep_caches

OrdinaryDiffEq._reshape

OrdinaryDiffEq._vec

OrdinaryDiffEq.add_saveat!

OrdinaryDiffEq.add_tstop!

OrdinaryDiffEq.addat!

OrdinaryDiffEq.addat_non_user_cache!

DiffEqBase.addsteps! — Function.

By default, Hermite interpolant so update the derivative at the two ends

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

Called to add the extra k9, k10, k11 steps for the Order 5 interpolation when needed

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

Called to add the extra k9, k10, k11 steps for the Order 5 interpolation when needed

OrdinaryDiffEq.alg_adaptive_order

OrdinaryDiffEq.alg_autodiff

OrdinaryDiffEq.alg_cache

OrdinaryDiffEq.alg_cache_expRK — Function.

alg_cache_expRK(alg,u,uEltypeNoUnits,uprev,f,t,dt,p,du1,tmp,dz,plist)

Construct the non-standard caches (not uType or rateType) for ExpRK integrators.

plist is a list of integers each corresponding to the order of a phiv(!) call in perform_step!.

OrdinaryDiffEq.alg_cache_exprb — Function.

alg_cache_exprb(alg,uEltypeNoUnits,uprev,f,t,p,du1,tmp,dz,plist)

Construct the non-standard caches (not uType or rateType) for Exprb integrators.

plist is a list of integers each corresponding to the order of a phiv(!) call in perform_step!.

OrdinaryDiffEq.alg_can_repeat_jac

OrdinaryDiffEq.alg_extrapolates

OrdinaryDiffEq.alg_maximum_order

OrdinaryDiffEq.alg_order

OrdinaryDiffEq.alg_stability_size

OrdinaryDiffEq.apply_step!

OrdinaryDiffEq.auto_dt_reset!

OrdinaryDiffEq.backward_diff!

OrdinaryDiffEq.beta1_default

OrdinaryDiffEq.beta2_default

OrdinaryDiffEq.build_grad_config

OrdinaryDiffEq.build_jac_config

OrdinaryDiffEq.calc_J — Function.

calc_J(integrator,cache,is_compos)

Interface for calculating the jacobian.

For constant caches, a new jacobian object is returned whereas for mutable caches cache.J is updated. In both cases, if integrator.f has a custom jacobian update function, then it will be called for the update. Otherwise, either ForwardDiff or finite difference will be used depending on the jac_config of the cache.

OrdinaryDiffEq.calc_J! — Function.

calc_J!(integrator,cache,is_compos)

Interface for calculating the jacobian.

For constant caches, a new jacobian object is returned whereas for mutable caches cache.J is updated. In both cases, if integrator.f has a custom jacobian update function, then it will be called for the update. Otherwise, either ForwardDiff or finite difference will be used depending on the jac_config of the cache.

OrdinaryDiffEq.calc_W!

OrdinaryDiffEq.calc_coeff!

OrdinaryDiffEq.calc_dt_propose!

OrdinaryDiffEq.calc_rosenbrock_differentiation!

OrdinaryDiffEq.calc_tderivative

OrdinaryDiffEq.calc_tderivative!

OrdinaryDiffEq.calculate_residuals — Function.

calculate_residuals(ũ, u₀, u₁, α, ρ)

Calculate element-wise residuals

\[\frac{ũ}{α+\max{|u₀|,|u₁|}*ρ}\]

calculate_residuals(u₀, u₁, α, ρ)

Calculate element-wise residuals

\[\frac{ũ}{α+\max{|u₀|,|u₁|}*ρ}\]

OrdinaryDiffEq.calculate_residuals! — Function.

calculate_residuals!(out, ũ, u₀, u₁, α, ρ)

Save element-wise residuals

\[\frac{ũ}{α+\max{|u₀|,|u₁|}*ρ}\]

in out.

calculate_residuals!(out, u₀, u₁, α, ρ)

Save element-wise residuals

\[\frac{ũ}{α+\max{|u₀|,|u₁|}*ρ}\]

in out.

OrdinaryDiffEq.change_t_via_interpolation!

OrdinaryDiffEq.check_error

OrdinaryDiffEq.check_keywords

OrdinaryDiffEq.choose_algorithm!

OrdinaryDiffEq.choosedeg! — Function.

choosedeg!(cache) -> nothing

Calculate ms[mdeg] (the degree of the Chebyshev polynomial) and cache.recind (the index of recurrence parameters for that degree), where recf[recind:(recind+ms[mdeg]-2)] are the μ,κ pairs for the mdeg degree method.

OrdinaryDiffEq.chooseη!

OrdinaryDiffEq.constructDormandPrince — Function.

constructDormandPrince()

Constructs the tableau object for the Dormand-Prince Order 4/5 method.

OrdinaryDiffEq.copyat_or_push!

OrdinaryDiffEq.current_extrapolant

OrdinaryDiffEq.current_extrapolant!

OrdinaryDiffEq.current_interpolant

OrdinaryDiffEq.current_interpolant!

OrdinaryDiffEq.deleteat!

OrdinaryDiffEq.deleteat_non_user_cache!

OrdinaryDiffEq.derivative

OrdinaryDiffEq.derivative!

OrdinaryDiffEq.determine_chunksize

OrdinaryDiffEq.dp5threaded_adaptiveloop

OrdinaryDiffEq.dp5threaded_denseloop

OrdinaryDiffEq.dp5threaded_loop1

OrdinaryDiffEq.dp5threaded_loop2

OrdinaryDiffEq.dp5threaded_loop3

OrdinaryDiffEq.dp5threaded_loop4

OrdinaryDiffEq.dp5threaded_loop5

OrdinaryDiffEq.dp5threaded_loop6

OrdinaryDiffEq.du_cache

OrdinaryDiffEq.dual_cache

OrdinaryDiffEq.eval

OrdinaryDiffEq.expRK_operators — Function.

expRK_operators(alg,dt,A) -> ops

Compute operator(s) for an ExpRK algorithm. dt is the time step and A is the matrix form of the linear operator (from either a linear problem or a SplitODEProblem). All ExpRK methods that use caching operators should implement this method.

OrdinaryDiffEq.expand_ϕ_and_ϕstar!

OrdinaryDiffEq.fix_dt_at_bounds!

OrdinaryDiffEq.fsal_typeof

OrdinaryDiffEq.full_cache

OrdinaryDiffEq.g_coefs!

OrdinaryDiffEq.gamma_default

OrdinaryDiffEq.get_chunksize

OrdinaryDiffEq.get_current_adaptive_order

OrdinaryDiffEq.get_current_alg_autodiff

OrdinaryDiffEq.get_current_alg_order

OrdinaryDiffEq.get_current_isfsal

OrdinaryDiffEq.get_dt

OrdinaryDiffEq.get_du

OrdinaryDiffEq.get_du!

OrdinaryDiffEq.get_proposed_dt

OrdinaryDiffEq.get_tmp_cache

OrdinaryDiffEq.handle_callback_modifiers!

OrdinaryDiffEq.handle_callbacks!

OrdinaryDiffEq.handle_discontinuities!

OrdinaryDiffEq.handle_dt!

OrdinaryDiffEq.handle_tstop!

OrdinaryDiffEq.has_exp

OrdinaryDiffEq.has_expmv

OrdinaryDiffEq.has_expmv!

OrdinaryDiffEq.has_ldiv

OrdinaryDiffEq.has_ldiv!

OrdinaryDiffEq.has_mul

OrdinaryDiffEq.has_mul!

OrdinaryDiffEq.hermite_interpolant — Function.

Hairer Norsett Wanner Solving Ordinary Differential Euations I - Nonstiff Problems Page 190

Herimte Interpolation, chosen if no other dispatch for ode_interpolant

Herimte Interpolation, chosen if no other dispatch for ode_interpolant

Herimte Interpolation, chosen if no other dispatch for ode_interpolant

Herimte Interpolation, chosen if no other dispatch for ode_interpolant

OrdinaryDiffEq.hermite_interpolant!

OrdinaryDiffEq.include

OrdinaryDiffEq.init

OrdinaryDiffEq.initialize!

OrdinaryDiffEq.initialize_callbacks!

OrdinaryDiffEq.intervals

OrdinaryDiffEq.is_constant

OrdinaryDiffEq.is_constant_cache

OrdinaryDiffEq.is_nordsieck_change_order

OrdinaryDiffEq.is_stiff

OrdinaryDiffEq.isdiscrete

OrdinaryDiffEq.isdtchangeable

OrdinaryDiffEq.isfsal

OrdinaryDiffEq.isimplicit

OrdinaryDiffEq.isinplace

OrdinaryDiffEq.islinear

OrdinaryDiffEq.ismultistep

OrdinaryDiffEq.issplit

OrdinaryDiffEq.jac_cache

OrdinaryDiffEq.jac_iter

OrdinaryDiffEq.jacobian

OrdinaryDiffEq.jacobian!

OrdinaryDiffEq.linear_interpolant — Function.

Linear Interpolation

OrdinaryDiffEq.linear_interpolant!

OrdinaryDiffEq.loopfooter!

OrdinaryDiffEq.loopheader!

OrdinaryDiffEq.maxeig!

OrdinaryDiffEq.modify_dt_for_tstops!

OrdinaryDiffEq.nlsolve_f

OrdinaryDiffEq.nlsolve_functional!

OrdinaryDiffEq.nordsieck_adjust!

OrdinaryDiffEq.nordsieck_adjust_order!

OrdinaryDiffEq.nordsieck_decrement_wait!

OrdinaryDiffEq.nordsieck_finalize!

OrdinaryDiffEq.nordsieck_prepare_next!

OrdinaryDiffEq.nordsieck_rescale!

OrdinaryDiffEq.nordsieck_rewind!

OrdinaryDiffEq.ode_addsteps!

OrdinaryDiffEq.ode_determine_initdt

OrdinaryDiffEq.ode_extrapolant

OrdinaryDiffEq.ode_extrapolant!

OrdinaryDiffEq.ode_interpolant — Function.

odeinterpolant and odeinterpolant! dispatch

OrdinaryDiffEq.ode_interpolant!

OrdinaryDiffEq.ode_interpolation — Function.

ode_interpolation(tvals,ts,timeseries,ks)

Get the value at tvals where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

ode_interpolation(tval::Number,ts,timeseries,ks)

Get the value at tval where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

OrdinaryDiffEq.ode_interpolation! — Function.

ode_interpolation(tvals,ts,timeseries,ks)

Get the value at tvals where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

ode_interpolation!(out,tval::Number,ts,timeseries,ks)

Get the value at tval where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

OrdinaryDiffEq.perform_predict!

OrdinaryDiffEq.perform_step!

OrdinaryDiffEq.qmax_default

OrdinaryDiffEq.qmin_default

OrdinaryDiffEq.qsteady_max_default

OrdinaryDiffEq.qsteady_min_default

OrdinaryDiffEq.rand_cache

OrdinaryDiffEq.ratenoise_cache

OrdinaryDiffEq.reeval_internals_due_to_modification!

OrdinaryDiffEq.reinit!

OrdinaryDiffEq.reinterpolate_history!

OrdinaryDiffEq.remake

OrdinaryDiffEq.reset_alg_dependent_opts! — Function.

If no user default, then this will change the default to the defaults for the second algorithm.

OrdinaryDiffEq.reset_fsal!

OrdinaryDiffEq.resize!

OrdinaryDiffEq.resize_non_user_cache!

OrdinaryDiffEq.save_idxsinitialize

OrdinaryDiffEq.savevalues!

OrdinaryDiffEq.set_abstol!

OrdinaryDiffEq.set_gamma!

OrdinaryDiffEq.set_proposed_dt!

OrdinaryDiffEq.set_reltol!

OrdinaryDiffEq.set_t!

OrdinaryDiffEq.set_u!

OrdinaryDiffEq.setη!

OrdinaryDiffEq.solution_endpoint_match_cur_integrator!

OrdinaryDiffEq.solve

OrdinaryDiffEq.solve!

OrdinaryDiffEq.ssp_coefficient — Function.

ssp_coefficient(alg)

Return the SSP coefficient of the ODE algorithm alg. If one time step of size dt with alg can be written as a convex combination of explicit Euler steps with step sizes cᵢ * dt, the SSP coefficient is the minimal value of 1/cᵢ.

Examples

julia> ssp_coefficient(SSPRK104())
6

OrdinaryDiffEq.step!

OrdinaryDiffEq.step_accept_controller!

OrdinaryDiffEq.step_reject_controller!

OrdinaryDiffEq.stepsize_and_order!

OrdinaryDiffEq.stepsize_controller!

OrdinaryDiffEq.stepsize_η!

OrdinaryDiffEq.stepsize_η₊₁!

OrdinaryDiffEq.stepsize_η₋₁!

OrdinaryDiffEq.terminate!

OrdinaryDiffEq.transfer_cache!

OrdinaryDiffEq.trivial_limiter!

OrdinaryDiffEq.tstop_saveat_disc_handling

OrdinaryDiffEq.tuples

OrdinaryDiffEq.u_cache

OrdinaryDiffEq.u_modified!

OrdinaryDiffEq.unwrap_alg

OrdinaryDiffEq.update_coefficients

OrdinaryDiffEq.update_coefficients!

OrdinaryDiffEq.update_nordsieck_vector!

OrdinaryDiffEq.user_cache

OrdinaryDiffEq.uses_uprev

OrdinaryDiffEq.warn_compat

OrdinaryDiffEq.γstar

OrdinaryDiffEq.γₖ

OrdinaryDiffEq.ϕ_and_ϕstar!

OrdinaryDiffEq.ϕ_np1!

OrdinaryDiffEq.∫₋₁⁰dx