Docstrings

DelayDiffEq.AB3

DelayDiffEq.AB4

DelayDiffEq.AB5

DelayDiffEq.ABDF2

DelayDiffEq.ABM32

DelayDiffEq.ABM43

DelayDiffEq.ABM54

DelayDiffEq.AN5

DelayDiffEq.AbstractAnalyticalProblem

DelayDiffEq.AbstractMethodOfStepsAlgorithm

DelayDiffEq.AffineDiffEqOperator

DelayDiffEq.AnalyticalProblem

DelayDiffEq.Anas5

DelayDiffEq.AutoDP5

DelayDiffEq.AutoSwitch

DelayDiffEq.AutoTsit5

DelayDiffEq.AutoVern6

DelayDiffEq.AutoVern7

DelayDiffEq.AutoVern8

DelayDiffEq.AutoVern9

DelayDiffEq.BS3

DelayDiffEq.BS5

DelayDiffEq.BVProblem

DelayDiffEq.CFRLDDRK64

DelayDiffEq.CNAB2

DelayDiffEq.CNLF2

DelayDiffEq.CallbackSet

DelayDiffEq.CalvoSanz4

DelayDiffEq.CandyRoz4

DelayDiffEq.CarpenterKennedy2N54

DelayDiffEq.Cash4

DelayDiffEq.CompositeAlgorithm

DelayDiffEq.ContinuousCallback

DelayDiffEq.DAEFunction

DelayDiffEq.DAEProblem

DelayDiffEq.DAESolution

DelayDiffEq.DDEFunction

DelayDiffEq.DDEIntegrator

DelayDiffEq.DDEProblem

DelayDiffEq.DEDataArray

DelayDiffEq.DEDataMatrix

DelayDiffEq.DEDataVector

DelayDiffEq.DEFAULT_LINSOLVE

DelayDiffEq.DESolution

DelayDiffEq.DGLDDRK73_C

DelayDiffEq.DGLDDRK84_C

DelayDiffEq.DGLDDRK84_F

DelayDiffEq.DP5

DelayDiffEq.DP5Threaded

DelayDiffEq.DP8

DelayDiffEq.DPRKN12

DelayDiffEq.DPRKN6

DelayDiffEq.DPRKN8

DelayDiffEq.DelayDiffEq

DelayDiffEq.DelayDiffEqAlgorithm

DelayDiffEq.DiffEqBase

DelayDiffEq.Discontinuity — Type.

Discontinuity(t, order::Int)

Object of discontinuity of order order at time t, i.e. discontinuity of orderth derivative at time t.

DelayDiffEq.DiscontinuityCallback — Type.

DiscontinuityCallback(lags, discontinuities::Vector{<:Discontinuity};
                      [interp_points::Int=10, abstol=1e-12, reltol=0])

Callback that tracks discontinuities that are propagated by dependent lags of the form (u,p,t) -> lag(u,p,t).

Hereby a number interp_points of interpolation points are used to first check for different signs of functions $f(t) = T + lag(u(t),p,t) - t$, where $T$ is time point of a previous discontinuity and $t$ is contained in the current time interval. This shows that the current time interval contains propagated discontinuities of which the exact time point is then determined by a root finding algorithm. The sign at the lower bound of the time interval, i.e. at tprev, is set to 0 with absolute tolerance abstol and relative tolerance reltol.

DelayDiffEq.DiscreteCallback

DelayDiffEq.DiscreteFunction

DelayDiffEq.DiscreteProblem

DelayDiffEq.DynamicalODEFunction

DelayDiffEq.DynamicalODEProblem

DelayDiffEq.EPIRK4s3A

DelayDiffEq.EPIRK4s3B

DelayDiffEq.EPIRK5P1

DelayDiffEq.EPIRK5P2

DelayDiffEq.EPIRK5s3

DelayDiffEq.ERKN4

DelayDiffEq.ERKN5

DelayDiffEq.ETD1

DelayDiffEq.ETD2

DelayDiffEq.ETDRK2

DelayDiffEq.ETDRK3

DelayDiffEq.ETDRK4

DelayDiffEq.EXPRB53s3

DelayDiffEq.Euler

DelayDiffEq.Exp4

DelayDiffEq.ExplicitRK

DelayDiffEq.Exprb32

DelayDiffEq.Exprb43

DelayDiffEq.Feagin10

DelayDiffEq.Feagin12

DelayDiffEq.Feagin14

DelayDiffEq.FunctionMap

DelayDiffEq.GRK4A

DelayDiffEq.GRK4T

DelayDiffEq.GenericIIF1

DelayDiffEq.GenericIIF2

DelayDiffEq.GenericImplicitEuler

DelayDiffEq.GenericTrapezoid

DelayDiffEq.HSLDDRK64

DelayDiffEq.Hairer4

DelayDiffEq.Hairer42

DelayDiffEq.Heun

DelayDiffEq.HistoryFunction — Type.

HistoryFunction(h, sol, integrator)

Wrap history function h, solution sol, and integrator integrator to create a common interface for retrieving values at any time point with varying accuracy.

Before the initial time point of solution sol values are calculated by history function h, for time points in the time span of sol interpolated values of sol are returned, and after the final time point of sol an inter- or extrapolation of the current state of integrator integrator is retrieved.

DelayDiffEq.HochOst4

DelayDiffEq.IMEXEuler

DelayDiffEq.IRKN3

DelayDiffEq.IRKN4

DelayDiffEq.ImplicitEuler

DelayDiffEq.ImplicitMidpoint

DelayDiffEq.JVODE

DelayDiffEq.JVODE_Adams

DelayDiffEq.JVODE_BDF

DelayDiffEq.KahanLi6

DelayDiffEq.KahanLi8

DelayDiffEq.KenCarp3

DelayDiffEq.KenCarp4

DelayDiffEq.KenCarp5

DelayDiffEq.Kvaerno3

DelayDiffEq.Kvaerno4

DelayDiffEq.Kvaerno5

DelayDiffEq.LawsonEuler

DelayDiffEq.LinSolveFactorize

DelayDiffEq.LinearExponential

DelayDiffEq.MEBDF2

DelayDiffEq.McAte2

DelayDiffEq.McAte3

DelayDiffEq.McAte4

DelayDiffEq.McAte42

DelayDiffEq.McAte5

DelayDiffEq.McAte8

DelayDiffEq.MethodOfSteps — Type.

MethodOfSteps(alg; constrained::Bool=false, fixedpoint_abstol=nothing,
              fixedpoint_reltol=nothing, fixedpoint_norm=nothing,
              max_fixedpoint_iters::Int=10)

Construct an algorithm that solves delay differential equations by the method of steps, where alg is an ODE algorithm from OrdinaryDiffEq.jl without lazy interpolation upon which the calculation of steps is based.

If the algorithm is constrained only steps of size at most the minimal delay will be taken. If it is unconstrained, fixed-point iteration is applied for step sizes that exceed the minimal delay.

The absolute and relative tolerance of the fixed-point iterations can be set by fixedpoint_abstol and fixedpoint_reltol, respectively, either as scalars or vectors. Based on these tolerances error estimates are calculated during the fixed-point iterations with a norm that may be specified as fixedpoint_norm. Fixed-point iterations are stopped if the error estimate is less than 1 or after the maximal number max_fixedpoint_iters of iteration steps.

DelayDiffEq.Midpoint

DelayDiffEq.MidpointSplitting

DelayDiffEq.MonteCarloProblem

DelayDiffEq.MonteCarloSolution

DelayDiffEq.MonteCarloSummary

DelayDiffEq.MonteCarloTestSolution

DelayDiffEq.NDBLSRK124

DelayDiffEq.NDBLSRK134

DelayDiffEq.NDBLSRK144

DelayDiffEq.NLAnderson

DelayDiffEq.NLFunctional

DelayDiffEq.NLNewton

DelayDiffEq.NoiseProblem

DelayDiffEq.NorsettEuler

DelayDiffEq.Nystrom4

DelayDiffEq.Nystrom4VelocityIndependent

DelayDiffEq.Nystrom5VelocityIndependent

DelayDiffEq.ODEFunction

DelayDiffEq.ODEProblem

DelayDiffEq.ODESolution

DelayDiffEq.ORK256

DelayDiffEq.OrdinaryDiffEq

DelayDiffEq.OrdinaryDiffEqAlgorithm

DelayDiffEq.OwrenZen3

DelayDiffEq.OwrenZen4

DelayDiffEq.OwrenZen5

DelayDiffEq.ParsaniKetchesonDeconinck3S105

DelayDiffEq.ParsaniKetchesonDeconinck3S173

DelayDiffEq.ParsaniKetchesonDeconinck3S184

DelayDiffEq.ParsaniKetchesonDeconinck3S205

DelayDiffEq.ParsaniKetchesonDeconinck3S32

DelayDiffEq.ParsaniKetchesonDeconinck3S53

DelayDiffEq.ParsaniKetchesonDeconinck3S82

DelayDiffEq.ParsaniKetchesonDeconinck3S94

DelayDiffEq.PseudoVerletLeapfrog

DelayDiffEq.QBDF

DelayDiffEq.QBDF1

DelayDiffEq.QBDF2

DelayDiffEq.QNDF

DelayDiffEq.QNDF1

DelayDiffEq.QNDF2

DelayDiffEq.RK4

DelayDiffEq.RK46NL

DelayDiffEq.RKC

DelayDiffEq.ROCK2

DelayDiffEq.ROCK4

DelayDiffEq.RODEFunction

DelayDiffEq.RODEProblem

DelayDiffEq.RODESolution

DelayDiffEq.ROS3P

DelayDiffEq.RadauIIA5

DelayDiffEq.Ralston

DelayDiffEq.RichardsonEuler

DelayDiffEq.Rodas3

DelayDiffEq.Rodas4

DelayDiffEq.Rodas42

DelayDiffEq.Rodas4P

DelayDiffEq.Rodas5

DelayDiffEq.Ros4LStab

DelayDiffEq.RosShamp4

DelayDiffEq.Rosenbrock23

DelayDiffEq.Rosenbrock32

DelayDiffEq.Ruth3

DelayDiffEq.SBDF2

DelayDiffEq.SBDF3

DelayDiffEq.SBDF4

DelayDiffEq.SDEFunction

DelayDiffEq.SDEProblem

DelayDiffEq.SDIRK2

DelayDiffEq.SSPRK104

DelayDiffEq.SSPRK22

DelayDiffEq.SSPRK33

DelayDiffEq.SSPRK432

DelayDiffEq.SSPRK53

DelayDiffEq.SSPRK53_2N1

DelayDiffEq.SSPRK53_2N2

DelayDiffEq.SSPRK54

DelayDiffEq.SSPRK63

DelayDiffEq.SSPRK73

DelayDiffEq.SSPRK83

DelayDiffEq.SSPRK932

DelayDiffEq.SSPRKMSVS32

DelayDiffEq.SSPRKMSVS43

DelayDiffEq.SSPSDIRK2

DelayDiffEq.SecondOrderODEProblem

DelayDiffEq.SofSpa10

DelayDiffEq.SplitEuler

DelayDiffEq.SplitFunction

DelayDiffEq.SplitODEProblem

DelayDiffEq.SplitSDEFunction

DelayDiffEq.SplitSDEProblem

DelayDiffEq.SteadyStateProblem

DelayDiffEq.SteadyStateSolution

DelayDiffEq.SymplecticEuler

DelayDiffEq.TRBDF2

DelayDiffEq.TSLDDRK74

DelayDiffEq.TanYam7

DelayDiffEq.TimeChoiceIterator

DelayDiffEq.Trapezoid

DelayDiffEq.Tsit5

DelayDiffEq.TsitPap8

DelayDiffEq.TwoPointBVProblem

DelayDiffEq.VCAB3

DelayDiffEq.VCAB4

DelayDiffEq.VCAB5

DelayDiffEq.VCABM

DelayDiffEq.VCABM3

DelayDiffEq.VCABM4

DelayDiffEq.VCABM5

DelayDiffEq.Veldd4

DelayDiffEq.Velds4

DelayDiffEq.VelocityVerlet

DelayDiffEq.VerletLeapfrog

DelayDiffEq.Vern6

DelayDiffEq.Vern7

DelayDiffEq.Vern8

DelayDiffEq.Vern9

DelayDiffEq.Yoshida6

DelayDiffEq.add_next_discontinuities! — Function.

add_next_discontinuities!(integrator::DDEIntegrator, order[, t=integrator.t])

Add discontinuities of next order that are propagated from discontinuity of order order at time t in integrator, but only if order is less or equal than the order of the applied method or the problem is neutral.

Discontinuities caused by constant delays are immediately calculated, and discontinuities caused by dependent delays are tracked by a callback.

DelayDiffEq.add_saveat!

DelayDiffEq.add_tstop!

DelayDiffEq.addat!

DelayDiffEq.addat_non_user_cache!

DelayDiffEq.addsteps!

DelayDiffEq.advance_ode_integrator! — Function.

advance_ode_integrator!(integrator::DDEIntegrator)

Advance ODE integrator of integrator to next time interval, values and complete interpolation data of integrator.

DelayDiffEq.agrees — Function.

agrees(h, u, p, t)

Determine whether history function evaluates to u at time point t for parameters p.

DelayDiffEq.assign_expr — Function.

assign_expr(::Val{name}, ::Type{T}, ::Type{cache})

Create expression that extracts field name of type T from cache of type cache to variable name.

Hereby u, uprev, uprev2, and function f are updated, if required.

DiffEqBase.auto_dt_reset! — Function.

auto_dt_reset!(dde_int::DDEIntegrator)

Automatically determine initial time step of dde_int.

DelayDiffEq.build_linked_cache — Function.

build_linked_cache(cache, alg, u, uprev, uprev2, f, t, dt)

Create cache for algorithm alg from existing cache cache with updated u, uprev, uprev2, f, t, and dt.

DelayDiffEq.build_solution_array — Function.

build_solution_array(integrator::DDEIntegrator)

Create a DiffEqArray of the time points and values that form the solution of integrator.

DelayDiffEq.build_solution_interpolation — Function.

build_solution_interpolation(integrator::DDEIntegrator, sol::DiffEqArray)

Create interpolation data to solution of integrator, which is formed by time points and values in sol.

DelayDiffEq.change_t_via_interpolation!

DelayDiffEq.check_error

DelayDiffEq.check_keywords

DelayDiffEq.constant_extrapolant — Function.

constant_extrapolant(t, integrator::DEIntegrator, idxs, deriv)

Calculate constant extrapolation of derivative deriv at time t and indices idxs for integrator.

DelayDiffEq.constant_extrapolant! — Function.

constant_extrapolant!(val, t, integrator::DEIntegrator, idxs, deriv)

Calculate constant extrapolation of derivative deriv at time t and indices idxs for integrator, and save result in val if val is not nothing.

DelayDiffEq.constructDormandPrince

DelayDiffEq.copyat_or_push!

DelayDiffEq.deleteat!

DelayDiffEq.deleteat_non_user_cache!

DelayDiffEq.determine_discontinuity_existence — Function.

determine_discontinuity_existence(prev_sign::Int, next_sign::Int, Θs, f)

Determine whether function f has a root in the interval [0, 1] by checking signs of f at 0 and 1 (prev_sign and next_sign, respectively) and at interpolation points in Θs.

This corresponds to the existence of a propagated discontinuity.

DelayDiffEq.du_cache

DelayDiffEq.eval

DelayDiffEq.find_discontinuity_time — Function.

find_discontinuity_time(integrator::DDEIntegrator, callback::DiscontinuityCallback,
                        prev_sign::Int, next_sign::Int, Θs, f)

Find time of propagated discontinuity for a certain dependent delay and previous discontinuity, which is root of the function f, in the current time interval of integrator with interpolation points Θs. Hereby f shows signs prev_sign and next_sign at both ends of the time interval.

DelayDiffEq.fsal_typeof — Function.

fsal_typeof(integrator::ODEIntegrator)

Return type of FSAL of integrator.

DelayDiffEq.full_cache

DelayDiffEq.get_dt

DelayDiffEq.get_du

DelayDiffEq.get_du!

DiffEqBase.get_proposed_dt — Function.

get_proposed_dt(integrator::DDEIntegrator)

Get the time step that integrator will take after the current step.

DelayDiffEq.get_tmp_cache

DelayDiffEq.has_exp

DelayDiffEq.has_expmv

DelayDiffEq.has_expmv!

DelayDiffEq.has_ldiv

DelayDiffEq.has_ldiv!

DelayDiffEq.has_mul

DelayDiffEq.has_mul!

DelayDiffEq.include

DelayDiffEq.init

DiffEqBase.initialize! — Function.

initialize!(integrator::DDEIntegrator)

Set initial values of integrator.

DelayDiffEq.intervals

DelayDiffEq.is_constant

DelayDiffEq.isconstrained

DelayDiffEq.isdiscrete

DelayDiffEq.isfsal

DelayDiffEq.isinplace

DelayDiffEq.ode_addsteps!

DelayDiffEq.ode_interpolant

DelayDiffEq.rand_cache

DelayDiffEq.ratenoise_cache

DelayDiffEq.reduce_solution! — Function.

reduce_solution!(integrator::DDEIntegrator, tmax)

Remove time points of ODE solution of integrator up to time tmax that are not required for calculation of DDE solution.

DiffEqBase.reeval_internals_due_to_modification! — Function.

reeval_internals_due_to_modification!(integrator::DDEIntegrator)

Recalculate interpolation data and update ODE integrator after changes by callbacks.

DiffEqBase.reinit! — Function.

reinit!(integrator::DDEIntegrator[, u0 = integrator.sol.prob.u0;
        t0 = integrator.sol.prob.tspan[1],
        tf = integrator.sol.prob.tspan[2],
        erase_sol = true,
        kwargs...])

Reinitialize integrator with (optionally) different initial state u0, different integration interval from t0 to tf, and erased solution if erase_sol = true.

DelayDiffEq.remake

DelayDiffEq.resize!

DelayDiffEq.resize_non_user_cache!

DelayDiffEq.savevalues!

DelayDiffEq.set_abstol!

DiffEqBase.set_proposed_dt! — Function.

set_proposed_dt!(integrator::DDEIntegrator, dt)

Set the time step that integrator will take after the current step to dt.

DelayDiffEq.set_reltol!

DelayDiffEq.set_t!

DelayDiffEq.set_u!

DelayDiffEq.solve

DelayDiffEq.solve!

DelayDiffEq.step!

DiffEqBase.terminate! — Function.

terminate!(integrator::DDEIntegrator)

Stop further calculations of integrator.

DelayDiffEq.tuples

DelayDiffEq.u_cache

DiffEqBase.u_modified! — Function.

u_modified!(integrator::DDEIntegrator, bool::Bool)

Signal integrator whether state vector u was modified by a callback.

A modified u will lead to recalculations in order to prevent discontinuities.

DelayDiffEq.update_coefficients

DelayDiffEq.update_coefficients!

DelayDiffEq.user_cache

DelayDiffEq.warn_compat