Docstrings

GitHub.@api_default — Macro.

For a method taking an API argument, add a new method without the API argument that just calls the method with DEFAULT_API.

GitHub.AnonymousAuth

GitHub.App

GitHub.Authorization

GitHub.Blob

GitHub.Branch

GitHub.COMMENT_EVENTS

GitHub.CheckRun

GitHub.Checks

GitHub.Comment

GitHub.CommentListener

GitHub.Commit

GitHub.Content

GitHub.DEFAULT_API

GitHub.ENTROPY

GitHub.EventListener

GitHub.Gist

GitHub.GitCommit

GitHub.GitHub

GitHub.GitHubAPI — Type.

Represents the API to interact with, either an actual GitHub instance, or a mock API for testing purposes

GitHub.GitHubType

GitHub.GitHubWebAPI

GitHub.Installation

GitHub.Issue

GitHub.JWTAuth

GitHub.OAuth2

GitHub.Owner

GitHub.PullRequest

GitHub.RNG

GitHub.Reference

GitHub.Repo

GitHub.Review

GitHub.Status

GitHub.Tag

GitHub.Team

GitHub.Tree

GitHub.UsernamePassAuth

GitHub.Webhook

GitHub.WebhookEvent

GitHub.__init__

GitHub.add_collaborator

GitHub.api_uri

GitHub.app

GitHub.authenticate

GitHub.authenticate_headers!

GitHub.base64_to_base64url

GitHub.blob

GitHub.branch

GitHub.branches

GitHub.build_content_response

GitHub.check_membership

GitHub.chopz

GitHub.close_pull_request

GitHub.collaborators

GitHub.comment

GitHub.comments

GitHub.commit

GitHub.commits

GitHub.content_uri

GitHub.contributors

GitHub.create_access_token

GitHub.create_blob

GitHub.create_check_run

GitHub.create_comment

GitHub.create_file

GitHub.create_fork

GitHub.create_gist

GitHub.create_gist_fork

GitHub.create_gitcommit

GitHub.create_issue

GitHub.create_pull_request

GitHub.create_reference

GitHub.create_repo

GitHub.create_status

GitHub.create_tag

GitHub.create_tree

GitHub.create_webhook

GitHub.delete_comment

GitHub.delete_file

GitHub.delete_gist

GitHub.delete_reference

GitHub.directory

GitHub.dismiss_review

GitHub.edit_comment

GitHub.edit_gist

GitHub.edit_issue

GitHub.eval

GitHub.event_from_payload!

GitHub.event_header

GitHub.extract_nullable

GitHub.extract_page_url

GitHub.file

GitHub.find_page_link

GitHub.followers

GitHub.following

GitHub.forks

GitHub.from_valid_repo

GitHub.get_page_links

GitHub.gh_delete

GitHub.gh_delete_json

GitHub.gh_get

GitHub.gh_get_json

GitHub.gh_get_paged_json

GitHub.gh_patch

GitHub.gh_patch_json

GitHub.gh_post

GitHub.gh_post_json

GitHub.gh_put

GitHub.gh_put_json

GitHub.gist

GitHub.gist_forks

GitHub.gists

GitHub.gitcommit

GitHub.github2json

GitHub.github_paged_get

GitHub.github_request

GitHub.handle_comment

GitHub.handle_event_request

GitHub.handle_response_error

GitHub.has_event_header

GitHub.has_page_links

GitHub.has_sig_header

GitHub.has_valid_secret

GitHub.include

GitHub.installations

GitHub.is_valid_event

GitHub.iscollaborator

GitHub.isorg

GitHub.issue

GitHub.issues

GitHub.json2github

GitHub.kind_err_str

GitHub.members

GitHub.name

GitHub.namefield

GitHub.orgs

GitHub.owner

GitHub.permalink

GitHub.prune_github_value

GitHub.pubkeys

GitHub.pull_request

GitHub.pull_requests

GitHub.rate_limit

GitHub.readme

GitHub.refereces

GitHub.reference

GitHub.references

GitHub.remove_collaborator

GitHub.reply_to

GitHub.repo

GitHub.repos

GitHub.reviews

GitHub.sig_header

GitHub.star

GitHub.star_gist

GitHub.stargazers

GitHub.starred

GitHub.starred_gists

GitHub.stats

GitHub.status

GitHub.statuses

GitHub.tag

GitHub.teams

GitHub.tree

GitHub.typprefix

GitHub.unstar

GitHub.unstar_gist

GitHub.unwatch

GitHub.unwrap_union_types

GitHub.update_check_run

GitHub.update_file

GitHub.update_pull_request

GitHub.update_reference

GitHub.users

GitHub.watch

GitHub.watched

GitHub.watchers