Docstrings

DiffEqBase.@CSI_str

DiffEqBase.@add_kwonly — Macro.

@add_kwonly function_definition

Define keyword-only version of the function_definition.

@add_kwonly function f(a, b; c=1, d=2)
    ...
end

expands to:

function f(x; y=1)
    ...
end
function f(; x = error("No argument x"), y=1)
    ...
end

DiffEqBase.@def

DiffEqBase.AbstractAnalyticalProblem

DiffEqBase.AbstractAnalyticalSolution

DiffEqBase.AbstractBVProblem

DiffEqBase.AbstractConstantLagDDEProblem

DiffEqBase.AbstractContinuousCallback

DiffEqBase.AbstractDAEAlgorithm

DiffEqBase.AbstractDAEFunction

DiffEqBase.AbstractDAEIntegrator

DiffEqBase.AbstractDAEProblem

DiffEqBase.AbstractDAESolution

DiffEqBase.AbstractDDEAlgorithm

DiffEqBase.AbstractDDEFunction

DiffEqBase.AbstractDDEIntegrator

DiffEqBase.AbstractDDEProblem

DiffEqBase.AbstractDDESolution

DiffEqBase.AbstractDEOptions

DiffEqBase.AbstractDiffEqFunction

DiffEqBase.AbstractDiffEqInterpolation

DiffEqBase.AbstractDiffEqLinearOperator

DiffEqBase.AbstractDiffEqOperator

DiffEqBase.AbstractDiscreteCallback

DiffEqBase.AbstractDiscreteFunction

DiffEqBase.AbstractDiscreteProblem

DiffEqBase.AbstractDynamicalODEProblem

DiffEqBase.AbstractMonteCarloEstimator

DiffEqBase.AbstractMonteCarloProblem

DiffEqBase.AbstractMonteCarloSolution

DiffEqBase.AbstractNoTimeSolution

DiffEqBase.AbstractNoiseProblem

DiffEqBase.AbstractNoiseProcess

DiffEqBase.AbstractODEAlgorithm

DiffEqBase.AbstractODEFunction

DiffEqBase.AbstractODEIntegrator

DiffEqBase.AbstractODEProblem

DiffEqBase.AbstractODESolution

DiffEqBase.AbstractParameterizedFunction

DiffEqBase.AbstractRODEAlgorithm

DiffEqBase.AbstractRODEFunction

DiffEqBase.AbstractRODEIntegrator

DiffEqBase.AbstractRODEProblem

DiffEqBase.AbstractRODESolution

DiffEqBase.AbstractReactionNetwork

DiffEqBase.AbstractSDEAlgorithm

DiffEqBase.AbstractSDEFunction

DiffEqBase.AbstractSDEIntegrator

DiffEqBase.AbstractSDEProblem

DiffEqBase.AbstractSecondOrderODEAlgorithm

DiffEqBase.AbstractSecondOrderODEIntegrator

DiffEqBase.AbstractSecondOrderODEProblem

DiffEqBase.AbstractSensitivitySolution

DiffEqBase.AbstractSplitODEProblem

DiffEqBase.AbstractSplitSDEProblem

DiffEqBase.AbstractSteadyStateAlgorithm

DiffEqBase.AbstractSteadyStateIntegrator

DiffEqBase.AbstractSteadyStateProblem

DiffEqBase.AbstractSteadyStateSolution

DiffEqBase.AbstractTimeseriesSolution

DiffEqBase.AffineDiffEqOperator — Type.

AffineDiffEqOperator{T} <: AbstractDiffEqOperator{T}

Ex: (A₁(t) + ... + Aₙ(t))*u + B₁(t) + ... + Bₙ(t)

AffineDiffEqOperator{T}(As,Bs,u_cache=nothing)

Takes in two tuples for split Affine DiffEqs

update_coefficients! works by updating the coefficients of the component operators.

Function calls L(u,p,t) and L(u,p,t,du) are fallbacks interpretted in this form. This will allow them to work directly in the nonlinear ODE solvers without modification.

f(u,p,t,du) is only allowed if a u_cache is given

B(t) can be Union{Number,AbstractArray}, in which case they are constants. Otherwise they are interpreted they are functions v=B(t) and B(v,t)

Solvers will see this operator from integrator.f and can interpret it by checking the internals of As and Bs. For example, it can check is_constant(As[1]) etc.

DiffEqBase.AnalyticalProblem

DiffEqBase.BVProblem

DiffEqBase.CallbackSet

DiffEqBase.ConstantInterpolation

DiffEqBase.ContinuousCallback

DiffEqBase.ConvergenceSetup

DiffEqBase.DAEFunction

DiffEqBase.DAEProblem

DiffEqBase.DAESolution

DiffEqBase.DDEFunction

DiffEqBase.DDEProblem

DiffEqBase.DEAlgorithm

DiffEqBase.DECache

DiffEqBase.DECallback

DiffEqBase.DECostFunction

DiffEqBase.DEDataArray

DiffEqBase.DEDataMatrix

DiffEqBase.DEDataVector

DiffEqBase.DEElement

DiffEqBase.DEFAULT_LINSOLVE

DiffEqBase.DEFAULT_PLOT_FUNC

DiffEqBase.DEIntegrator

DiffEqBase.DEProblem

DiffEqBase.DESensitivity

DiffEqBase.DESolution

DiffEqBase.DESolutionIterator

DiffEqBase.DISCRETE_INPLACE_DEFAULT

DiffEqBase.DISCRETE_OUTOFPLACE_DEFAULT

DiffEqBase.DiffEqBase

DiffEqBase.DiscreteCallback

DiffEqBase.DiscreteFunction

DiffEqBase.DiscreteProblem

DiffEqBase.DynamicalODEFunction

DiffEqBase.DynamicalODEProblem

DiffEqBase.ExplicitRKTableau — Type.

ExplicitRKTableau

Holds a tableau which defines an explicit Runge-Kutta method.

DiffEqBase.HermiteInterpolation

DiffEqBase.INITIALIZE_DEFAULT

DiffEqBase.ImplicitRKTableau — Type.

ImplicitRKTableau

Holds a tableau which defines an implicit Runge-Kutta method.

DiffEqBase.IntegratorIntervals

DiffEqBase.IntegratorTuples

DiffEqBase.InternalEuler

DiffEqBase.LinSolveFactorize

DiffEqBase.LinearInterpolation

DiffEqBase.MonteCarloProblem

DiffEqBase.MonteCarloSolution

DiffEqBase.MonteCarloSummary

DiffEqBase.MonteCarloTestSolution

DiffEqBase.NO_COLOR

DiffEqBase.NoiseProblem

DiffEqBase.ODEFunction

DiffEqBase.ODEProblem

DiffEqBase.ODERKTableau

DiffEqBase.ODESolution

DiffEqBase.ODE_DEFAULT_ISOUTOFDOMAIN

DiffEqBase.ODE_DEFAULT_NORM

DiffEqBase.ODE_DEFAULT_PROG_MESSAGE

DiffEqBase.ODE_DEFAULT_UNSTABLE_CHECK

DiffEqBase.RECOMPILE_BY_DEFAULT

DiffEqBase.RODEFunction

DiffEqBase.RODEProblem

DiffEqBase.RODESolution

DiffEqBase.SDEFunction

DiffEqBase.SDEProblem

DiffEqBase.SecondOrderODEProblem

DiffEqBase.SplitFunction

DiffEqBase.SplitODEProblem

DiffEqBase.SplitSDEFunction

DiffEqBase.SplitSDEProblem

DiffEqBase.StandardBVProblem

DiffEqBase.StandardODEProblem

DiffEqBase.StandardSDEProblem

DiffEqBase.SteadyStateProblem

DiffEqBase.SteadyStateSolution

DiffEqBase.TYPE_COLOR

DiffEqBase.Tableau

DiffEqBase.TimeChoiceIterator

DiffEqBase.TwoPointBVPFunction

DiffEqBase.TwoPointBVProblem

DiffEqBase.UNITLESS_ABS2

DiffEqBase.__has_analytic

DiffEqBase.__has_invW

DiffEqBase.__has_invW_t

DiffEqBase.__has_jac

DiffEqBase.__has_paramderiv

DiffEqBase.__has_paramjac

DiffEqBase.__has_syms

DiffEqBase.__has_tgrad

DiffEqBase.__init

DiffEqBase.__solve

DiffEqBase.add_analytic_labels!

DiffEqBase.add_kwonly

DiffEqBase.add_labels!

DiffEqBase.add_saveat!

DiffEqBase.add_tstop!

DiffEqBase.addat!

DiffEqBase.addat_non_user_cache!

DiffEqBase.auto_dt_reset!

DiffEqBase.build_solution

DiffEqBase.calculate_monte_errors

DiffEqBase.calculate_solution_errors!

DiffEqBase.check_error — Function.

check_error(integrator)

Check state of integrator and return one of the Return Codes

DiffEqBase.check_error! — Function.

check_error!(integrator)

Same as check_error but also set solution's return code (integrator.sol.retcode) and run postamble!.

DiffEqBase.check_first_arg

DiffEqBase.check_keywords

DiffEqBase.copy_fields — Function.

copy_fields(arr:AbstractArray, template::DEDataArray)

Create DEDataArray that wraps arr with all other fields set to a deep copy of the value in template.

DiffEqBase.copy_fields! — Function.

copy_fields!(dest::T, src::T) where {T<:DEDataArray}

Replace all fields of dest except of its wrapped array with a copy of the value in src. Arrays are recursively copied.

DiffEqBase.deleteat!

DiffEqBase.deleteat_non_user_cache!

DiffEqBase.done

DiffEqBase.du_cache

DiffEqBase.eval

DiffEqBase.expmv

DiffEqBase.expmv!

DiffEqBase.full_cache

DiffEqBase.get_concrete_problem

DiffEqBase.get_dt

DiffEqBase.get_du

DiffEqBase.get_du!

DiffEqBase.get_proposed_dt

DiffEqBase.get_tmp_cache

DiffEqBase.handle_distribution_u0

DiffEqBase.has_analytic

DiffEqBase.has_exp

DiffEqBase.has_expmv

DiffEqBase.has_expmv!

DiffEqBase.has_invW

DiffEqBase.has_invW_t

DiffEqBase.has_jac

DiffEqBase.has_ldiv

DiffEqBase.has_ldiv!

DiffEqBase.has_mul

DiffEqBase.has_mul!

DiffEqBase.has_paramjac

DiffEqBase.has_reinit

DiffEqBase.has_syms

DiffEqBase.has_tgrad

DiffEqBase.include

DiffEqBase.init

DiffEqBase.initialize!

DiffEqBase.interp_summary

DiffEqBase.interpolant — Function.

Hairer Norsett Wanner Solving Ordinary Differential Equations I - Nonstiff Problems Page 190

Hermite Interpolation

Hermite Interpolation

Hermite Interpolation

Hermite Interpolation

Linear Interpolation

Constant Interpolation

DiffEqBase.interpolant! — Function.

Hairer Norsett Wanner Solving Ordinary Differential Euations I - Nonstiff Problems Page 190

Hermite Interpolation

Hermite Interpolation

Hermite Interpolation

Hermite Interpolation

Linear Interpolation

Linear Interpolation

Constant Interpolation

Constant Interpolation

DiffEqBase.interpolation — Function.

interpolation(tval::Number,t,u,ks)

Get the value at tval where the solution is known at the times t (sorted), with values u and derivatives ks

DiffEqBase.interpolation! — Function.

interpolation(tvals,t,u,ks)

Get the value at tvals where the solution is known at the times t (sorted), with values u and derivatives ks

interpolation!(out,tval::Number,t,u,ks)

Get the value at tval where the solution is known at the times t (sorted), with values u and derivatives ks

DiffEqBase.interpret_vars

DiffEqBase.intervals

DiffEqBase.is_constant

DiffEqBase.is_diagonal_noise

DiffEqBase.isadaptive

DiffEqBase.isautodifferentiable

DiffEqBase.isinplace

DiffEqBase.last_step_failed

DiffEqBase.next

DiffEqBase.num_types_in_tuple

DiffEqBase.numargs — Function.

numparameters(f)

Returns the number of parameters of f for the method which has the most parameters.

DiffEqBase.parameterless_type

DiffEqBase.plot_indices

DiffEqBase.postamble!

DiffEqBase.problem_new_parameters

DiffEqBase.promote_tspan

DiffEqBase.reinit!

DiffEqBase.remake — Function.

remake(thing; <keyword arguments>)

Re-construct thing with new field values specified by the keyword arguments.

DiffEqBase.resize!

DiffEqBase.resize_non_user_cache!

DiffEqBase.savevalues!

DiffEqBase.set_abstol!

DiffEqBase.set_proposed_dt!

DiffEqBase.set_reltol!

DiffEqBase.set_t! — Function.

set_t!(integrator::DEIntegrator, t::Real)

Set current time point of the integrator to t.

DiffEqBase.set_u! — Function.

set_u!(integrator::DEIntegrator, u)

Set current state of the integrator to u.

DiffEqBase.set_ut! — Function.

set_ut!(integrator::DEIntegrator, u, t)

Set current state of the integrator to u and t

DiffEqBase.solplot_vecs_and_labels

DiffEqBase.solution_new_retcode

DiffEqBase.solution_new_tslocation

DiffEqBase.solve

DiffEqBase.solve!

DiffEqBase.split_callbacks

DiffEqBase.start

DiffEqBase.step! — Function.

step!(integ::DEIntegrator [, dt [, stop_at_tdt]])

Perform one (successful) step on the integrator.

Alternative, if a dt is given, then step! the integrator until there is a temporal difference ≥ dt in integ.t. When true is passed to the optional third argument, the integrator advances exactly dt.

DiffEqBase.struct_as_dict

DiffEqBase.terminate!

DiffEqBase.tuples

DiffEqBase.u_cache

DiffEqBase.u_modified!

DiffEqBase.u_n

DiffEqBase.undefined_exports — Function.

undefined_exports(mod)

List symbols export'ed but not actually defined.

DiffEqBase.update_coefficients

DiffEqBase.update_coefficients!

DiffEqBase.user_cache

DiffEqBase.warn_compat