Docstrings

StanMCMCChain.@animate

StanMCMCChain.@colorant_str

StanMCMCChain.@gif

StanMCMCChain.@layout

StanMCMCChain.@load

StanMCMCChain.@recipe

StanMCMCChain.@save

StanMCMCChain.@series

StanMCMCChain.@shorthands

StanMCMCChain.@userplot

StanMCMCChain.ABGR

StanMCMCChain.ADIN99

StanMCMCChain.ADIN99d

StanMCMCChain.ADIN99o

StanMCMCChain.AGray

StanMCMCChain.AGray32

StanMCMCChain.AHSI

StanMCMCChain.AHSL

StanMCMCChain.AHSV

StanMCMCChain.ALCHab

StanMCMCChain.ALCHuv

StanMCMCChain.ALMS

StanMCMCChain.ALab

StanMCMCChain.ALuv

StanMCMCChain.ARGB

StanMCMCChain.ARGB32

StanMCMCChain.AXYZ

StanMCMCChain.AYCbCr

StanMCMCChain.AYIQ

StanMCMCChain.AbstractBackend

StanMCMCChain.AbstractGray

StanMCMCChain.AbstractLayout

StanMCMCChain.AbstractPlot

StanMCMCChain.AbstractRGB

StanMCMCChain.AlphaColor

StanMCMCChain.Animation

StanMCMCChain.AxyY

StanMCMCChain.BGR

StanMCMCChain.BGRA

StanMCMCChain.BezierCurve

StanMCMCChain.CIE1931JV_CMF

StanMCMCChain.CIE1931J_CMF

StanMCMCChain.CIE1931_CMF

StanMCMCChain.CIE1964_CMF

StanMCMCChain.CMDSTAN_HOME

StanMCMCChain.Chains

StanMCMCChain.CmdStan

StanMCMCChain.Color

StanMCMCChain.Color3

StanMCMCChain.ColorAlpha

StanMCMCChain.ColorGradient

StanMCMCChain.ColorTypes

StanMCMCChain.Colorant

StanMCMCChain.ColorantNormed

StanMCMCChain.Colors

StanMCMCChain.DE_2000

StanMCMCChain.DE_94

StanMCMCChain.DE_AB

StanMCMCChain.DE_BFD

StanMCMCChain.DE_CMC

StanMCMCChain.DE_DIN99

StanMCMCChain.DE_DIN99d

StanMCMCChain.DE_DIN99o

StanMCMCChain.DE_JPC79

StanMCMCChain.DIN99

StanMCMCChain.DIN99A

StanMCMCChain.DIN99d

StanMCMCChain.DIN99dA

StanMCMCChain.DIN99o

StanMCMCChain.DIN99oA

StanMCMCChain.Diagnose

StanMCMCChain.Formatted

StanMCMCChain.Fractional

StanMCMCChain.GR

StanMCMCChain.Gray

StanMCMCChain.Gray24

StanMCMCChain.GrayA

StanMCMCChain.HSB

StanMCMCChain.HSI

StanMCMCChain.HSIA

StanMCMCChain.HSL

StanMCMCChain.HSLA

StanMCMCChain.HSV

StanMCMCChain.HSVA

StanMCMCChain.JLD

StanMCMCChain.KW

StanMCMCChain.LCHab

StanMCMCChain.LCHabA

StanMCMCChain.LCHuv

StanMCMCChain.LCHuvA

StanMCMCChain.LMS

StanMCMCChain.LMSA

StanMCMCChain.Lab

StanMCMCChain.LabA

StanMCMCChain.Luv

StanMCMCChain.LuvA

StanMCMCChain.MCMCChain

StanMCMCChain.MSC

StanMCMCChain.OHLC

StanMCMCChain.Optimize

StanMCMCChain.PlotThemes

StanMCMCChain.PlotUtils

StanMCMCChain.Plots

StanMCMCChain.RGB

StanMCMCChain.RGB1

StanMCMCChain.RGB24

StanMCMCChain.RGB4

StanMCMCChain.RGBA

StanMCMCChain.RecipeData

StanMCMCChain.RecipesBase

StanMCMCChain.Sample

StanMCMCChain.Segments

StanMCMCChain.Shape

StanMCMCChain.StanMCMCChain

StanMCMCChain.Stanmodel

StanMCMCChain.Statistics

StanMCMCChain.Surface

StanMCMCChain.Transparent3

StanMCMCChain.TransparentColor

StanMCMCChain.TransparentGray

StanMCMCChain.TransparentRGB

StanMCMCChain.Variational

StanMCMCChain.XYZ

StanMCMCChain.XYZA

StanMCMCChain.YCbCr

StanMCMCChain.YCbCrA

StanMCMCChain.YIQ

StanMCMCChain.YIQA

StanMCMCChain.adapted_grid

StanMCMCChain.add_backend

StanMCMCChain.add_theme

StanMCMCChain.addrequire

StanMCMCChain.aliases

StanMCMCChain.alpha

StanMCMCChain.alphacolor

StanMCMCChain.animate

StanMCMCChain.annotate!

StanMCMCChain.arrow

StanMCMCChain.attr!

StanMCMCChain.autcorplot

StanMCMCChain.autocorplot

StanMCMCChain.autocorplot!

StanMCMCChain.backend

StanMCMCChain.backend_name

StanMCMCChain.backend_object

StanMCMCChain.backends

StanMCMCChain.bar

StanMCMCChain.bar!

StanMCMCChain.barh

StanMCMCChain.barh!

StanMCMCChain.barhist

StanMCMCChain.barhist!

StanMCMCChain.base_color_type

StanMCMCChain.base_colorant_type

StanMCMCChain.bbox

StanMCMCChain.blue

StanMCMCChain.boxplot

StanMCMCChain.boxplot!

StanMCMCChain.brush

StanMCMCChain.ccolor

StanMCMCChain.center

StanMCMCChain.cgrad

StanMCMCChain.cgradients

StanMCMCChain.cie_color_match

StanMCMCChain.clibraries

StanMCMCChain.clibrary

StanMCMCChain.closeall

StanMCMCChain.color

StanMCMCChain.color_type

StanMCMCChain.coloralpha

StanMCMCChain.colordiff

StanMCMCChain.colormap

StanMCMCChain.colormatch

StanMCMCChain.comp1

StanMCMCChain.comp2

StanMCMCChain.comp3

StanMCMCChain.contour

StanMCMCChain.contour!

StanMCMCChain.contour3d

StanMCMCChain.contour3d!

StanMCMCChain.contourf

StanMCMCChain.contourf!

CmdStan.convert_a3d — Function.

convert_a3d

Convert the output file created by cmdstan to the shape of choice. Currently .

Method

convert_a3d(a3d_array, cnames, ::Val{Symbol})

Required arguments

* `a3d_array::Array{Float64}(n_draws, n_variables, n_chains`      : Read in from output files created by cmdstan                                   
* `cnames::Vector{AbstractString}`                                                 : Monitored variable names
* `::Val{Symbol}`                                                                             : Output format

Method called is based on the output_format defined in the stanmodel, e.g.:

   stanmodel = Stanmodel(num_samples=1200, thin=2, name="bernoulli", 
   model=bernoullimodel, output_format=:array);

Current formats supported are:

1. :array (a3d_array format, the default for CmdStan)
2. :dataFrame (DataFrame)
3. :mambachains (Mamba.Chains object)
4. :mcmcchain (TuringLang/Chains object)

Options 2 through 4 are respectively provided by the packages StanDataFrames, StanMamba and StanMCMCChain.

Return values

* `res`                       : Draws converted to the specified format.

StanMCMCChain.coords

StanMCMCChain.cor

StanMCMCChain.corner

StanMCMCChain.corner!

StanMCMCChain.cov

StanMCMCChain.creator

StanMCMCChain.current

StanMCMCChain.curve_points

StanMCMCChain.curves

StanMCMCChain.curves!

StanMCMCChain.cvec

StanMCMCChain.default

StanMCMCChain.default_cgrad

StanMCMCChain.density

StanMCMCChain.density!

StanMCMCChain.densityplot

StanMCMCChain.describe

StanMCMCChain.deuteranopic

StanMCMCChain.discretediag

StanMCMCChain.distinguishable_colors

StanMCMCChain.diverging_palette

StanMCMCChain.eval

StanMCMCChain.font

StanMCMCChain.frame

StanMCMCChain.gamutmax

StanMCMCChain.gamutmin

StanMCMCChain.gelmandiag

StanMCMCChain.get_color_palette

StanMCMCChain.getindex

StanMCMCChain.gewekediag

StanMCMCChain.gif

StanMCMCChain.gr

StanMCMCChain.gray

StanMCMCChain.green

StanMCMCChain.grid

StanMCMCChain.gui

StanMCMCChain.hdf5

StanMCMCChain.heatmap

StanMCMCChain.heatmap!

StanMCMCChain.heideldiag

StanMCMCChain.hex

StanMCMCChain.hexbin

StanMCMCChain.hexbin!

StanMCMCChain.histogram

StanMCMCChain.histogram!

StanMCMCChain.histogram2d

StanMCMCChain.histogram2d!

StanMCMCChain.histogramplot

StanMCMCChain.hline

StanMCMCChain.hline!

StanMCMCChain.hspan

StanMCMCChain.hspan!

StanMCMCChain.include

StanMCMCChain.inline

StanMCMCChain.inspectdr

StanMCMCChain.invisible

StanMCMCChain.isdark

StanMCMCChain.ismmappable

StanMCMCChain.iter_segments

StanMCMCChain.jldopen

StanMCMCChain.load

StanMCMCChain.mapc

StanMCMCChain.mapreducec

StanMCMCChain.mean

StanMCMCChain.mean!

StanMCMCChain.meanplot

StanMCMCChain.meanplot!

StanMCMCChain.median

StanMCMCChain.median!

StanMCMCChain.middle

StanMCMCChain.mixeddensity

StanMCMCChain.mixeddensity!

StanMCMCChain.mixeddensityplot

StanMCMCChain.mov

StanMCMCChain.mp4

StanMCMCChain.o_delete

StanMCMCChain.ohlc

StanMCMCChain.ohlc!

StanMCMCChain.optimize_datetime_ticks

StanMCMCChain.optimize_ticks

StanMCMCChain.palette

StanMCMCChain.path3d

StanMCMCChain.path3d!

StanMCMCChain.pgfplots

StanMCMCChain.pie

StanMCMCChain.pie!

StanMCMCChain.plot

StanMCMCChain.plot!

StanMCMCChain.plot3d

StanMCMCChain.plot3d!

StanMCMCChain.plot_color

StanMCMCChain.plotarea

StanMCMCChain.plotattr

StanMCMCChain.plotly

StanMCMCChain.plotlyjs

StanMCMCChain.plots_heatmap

StanMCMCChain.plots_heatmap!

StanMCMCChain.png

StanMCMCChain.portfoliocomposition

StanMCMCChain.portfoliocomposition!

StanMCMCChain.protanopic

StanMCMCChain.pyplot

StanMCMCChain.quantile

StanMCMCChain.quantile!

StanMCMCChain.quiver

StanMCMCChain.quiver!

StanMCMCChain.rafterydiag

StanMCMCChain.readmmap

StanMCMCChain.red

StanMCMCChain.reducec

StanMCMCChain.rgb_string

StanMCMCChain.rgba_string

StanMCMCChain.rotate

StanMCMCChain.rotate!

StanMCMCChain.save

StanMCMCChain.savefig

StanMCMCChain.scatter

StanMCMCChain.scatter!

StanMCMCChain.scatter3d

StanMCMCChain.scatter3d!

StanMCMCChain.scatterhist

StanMCMCChain.scatterhist!

StanMCMCChain.sequential_palette

StanMCMCChain.set_cmdstan_home!

StanMCMCChain.set_theme

StanMCMCChain.setindex!

StanMCMCChain.shape_coords

StanMCMCChain.showgradient

StanMCMCChain.showgradient!

StanMCMCChain.showlibrary

StanMCMCChain.showlibrary!

StanMCMCChain.showtheme

StanMCMCChain.showtheme!

StanMCMCChain.spy

StanMCMCChain.spy!

StanMCMCChain.stan

StanMCMCChain.std

StanMCMCChain.stdm

StanMCMCChain.stephist

StanMCMCChain.stephist!

StanMCMCChain.sticks

StanMCMCChain.sticks!

StanMCMCChain.stroke

StanMCMCChain.surface

StanMCMCChain.surface!

StanMCMCChain.test_examples

StanMCMCChain.text

StanMCMCChain.theme

StanMCMCChain.title!

StanMCMCChain.traceplot

StanMCMCChain.traceplot!

StanMCMCChain.translate

StanMCMCChain.translate!

StanMCMCChain.tritanopic

StanMCMCChain.truncate_module_path

StanMCMCChain.twinx

StanMCMCChain.unicodeplots

StanMCMCChain.var

StanMCMCChain.varm

StanMCMCChain.violin

StanMCMCChain.violin!

StanMCMCChain.vline

StanMCMCChain.vline!

StanMCMCChain.vspan

StanMCMCChain.vspan!

StanMCMCChain.weighted_color_mean

StanMCMCChain.whitebalance

StanMCMCChain.wireframe

StanMCMCChain.wireframe!

StanMCMCChain.with

StanMCMCChain.wrap

StanMCMCChain.xaxis!

StanMCMCChain.xflip!

StanMCMCChain.xgrid!

StanMCMCChain.xlabel!

StanMCMCChain.xlims

StanMCMCChain.xlims!

StanMCMCChain.xticks!

StanMCMCChain.xyY

StanMCMCChain.xyYA

StanMCMCChain.yaxis!

StanMCMCChain.yflip!

StanMCMCChain.ygrid!

StanMCMCChain.ylabel!

StanMCMCChain.ylims

StanMCMCChain.ylims!

StanMCMCChain.yticks!

StanMCMCChain.zlims

StanMCMCChain.zlims!