Docstrings

DiffEqBiological.@reaction_func

DiffEqBiological.@reaction_network — Macro.

Macro that inputs an expression corresponding to a reaction network and output a Reaction Network Structure that can be used as input to generation of SDE and ODE and Jump problems. Most arrows accepted (both right, left and bi drectional arrows). Note that while –> is a correct arrow, neither <– nor <–> works. Using non-filled arrows (⇐, ⟽, ⇒, ⟾, ⇔, ⟺) will disable mass kinetics and lets you cutomize reaction rates yourself. Use 0 or ∅ for degradation/creation to/from nothing. Example systems: ### Basic Usage ### rn = @reaction_network rType begin #Creates a reaction network of type rType. 2.0, X + Y –> XY #This will have reaction rate corresponding to 2.0*[X][Y] 2.0, XY ← X + Y #Identical to 2.0, X + Y –> XY end

### Manipulating Reaction Rates ###
rn = @reaction_network rType begin
    2.0, X + Y ⟾ XY                   #Ignores mass kinetics. This will have reaction rate corresponding to 2.0.
    2.0X, X + Y --> XY                 #Reaction rate needs not be constant. This will have reaction rate corresponding to 2.0*[X]*[X]*[Y].
    XY+log(X)^2, X + Y --> XY          #Reaction rate accepts quite complicated expressions (user defined functions must first be registered using the @reaction_func macro).
    hill(XY,2,2,2), X + Y --> XY       #Reaction inis activated by XY according to a hill function. hill(x,v,K,N).
    mm(XY,2,2), X + Y --> XY           #Reaction inis activated by XY according to a michaelis menten function. mm(x,v,K).
end

### Multipple Reactions on a Single Line ###
rn = @reaction_network rType begin
    (2.0,1.0), X + Y ↔ XY              #Identical to reactions (2.0, X + Y --> XY) and (1.0, XY --> X + Y).
    2.0, (X,Y) --> 0                   #This corresponds to both X and Y degrading at rate 2.0.
    (2.0, 1.0), (X,Y) --> 0            #This corresponds to X and Y degrading at rates 2.0 and 1.0, respectively.
    2.0, (X1,Y1) --> (X2,Y2)           #X1 and Y1 becomes X2 and Y2, respectively, at rate 2.0.
end

### Adding Parameters ###
kB = 2.0; kD = 1.0
p = [kB, kD]
p = []
rn = @reaction_network type begin
    (kB, kD), X + Y ↔ XY            #Lets you define parameters outside on network. Parameters can be changed without recalling the network.
end kB, kD

### Defining New Functions ###
@reaction_func my_hill_repression(x, v, k, n) = v*k^n/(k^n+x^n)     #Creates and adds a new function that the @reaction_network macro can see.
r = @reaction_network MyReactionType begin
    my_hill_repression(x, v_x, k_x, n_x), 0 --> x                       #After it has been added in @reaction_func the function can be used when defining new reaction networks.
end v_x k_x n_x

### Simulating Reaction Networks ###
probODE = ODEProblem(rn, args...; kwargs...)        #Using multiple dispatch the reaction network can be used as input to create ODE, SDE and Jump problems.
probSDE = SDEProblem(rn, args...; kwargs...)
probJump = JumpProblem(prob,aggregator::Direct,rn)

DiffEqBiological.BracketData

DiffEqBiological.ConstantRateJump

DiffEqBiological.DiffEqBiological

DiffEqBiological.DiffEqJump

DiffEqBiological.Direct

DiffEqBiological.DirectCR

DiffEqBiological.DirectFW

DiffEqBiological.ExtendedJumpArray

DiffEqBiological.FRM

DiffEqBiological.FRMFW

DiffEqBiological.JumpProblem

DiffEqBiological.JumpSet

DiffEqBiological.MassActionJump

DiffEqBiological.NRM

DiffEqBiological.RSSA

DiffEqBiological.ReactantStruct

DiffEqBiological.ReactionStruct

DiffEqBiological.RegularJump

DiffEqBiological.RegularSSA

DiffEqBiological.SDEProblem

DiffEqBiological.SSAStepper

DiffEqBiological.SimpleTauLeaping

DiffEqBiological.SortingDirect

DiffEqBiological.SplitCoupledJumpProblem

DiffEqBiological.SteadyStateProblem

DiffEqBiological.VariableRateJump

DiffEqBiological.add_reactants!

DiffEqBiological.bwd_arrows

DiffEqBiological.calculate_jac

DiffEqBiological.coordinate

DiffEqBiological.depgraph_from_network

DiffEqBiological.double_arrows

DiffEqBiological.empty_set

DiffEqBiological.eval

DiffEqBiological.expr_arr_to_block

DiffEqBiological.funcdict

DiffEqBiological.fwd_arrows

DiffEqBiological.get_f

DiffEqBiological.get_g

DiffEqBiological.get_jumps

DiffEqBiological.get_net_stoich

DiffEqBiological.get_num_majumps

DiffEqBiological.get_parameters

DiffEqBiological.get_reactants

DiffEqBiological.get_reactions

DiffEqBiological.get_stoch_diff

DiffEqBiological.get_substrate_stoich

DiffEqBiological.get_tup_arg

DiffEqBiological.hill

DiffEqBiological.hill_name

DiffEqBiological.include

DiffEqBiological.init

DiffEqBiological.make_func

DiffEqBiological.make_majump

DiffEqBiological.maketype

DiffEqBiological.mass_rate_DE

DiffEqBiological.mass_rate_SSA

DiffEqBiological.mm

DiffEqBiological.mm_name

DiffEqBiological.needs_depgraph

DiffEqBiological.needs_vartojumps_map

DiffEqBiological.network_to_jumpset

DiffEqBiological.no_mass_arrows

DiffEqBiological.push_reactions

DiffEqBiological.rate_to_indices

DiffEqBiological.recursive_clean!

DiffEqBiological.recursive_contains

DiffEqBiological.recursive_content

DiffEqBiological.recursive_replace!

DiffEqBiological.replace_names

DiffEqBiological.rxidxs_to_jidxs_map

DiffEqBiological.solve

DiffEqBiological.solve!

DiffEqBiological.spec_to_dep_jumps_map

DiffEqBiological.species_to_indices

DiffEqBiological.tup_leng

DiffEqBiological.update_reaction_info