Docstrings

BioSymbols.AA_A — Constant.

Alanine

BioSymbols.AA_B — Constant.

Aspartic Acid or Asparagine

BioSymbols.AA_C — Constant.

Cysteine

BioSymbols.AA_D — Constant.

Aspartic Acid

BioSymbols.AA_E — Constant.

Glutamic Acid

BioSymbols.AA_F — Constant.

Phenylalanine

BioSymbols.AA_G — Constant.

Glycine

BioSymbols.AA_Gap — Constant.

Amino Acid Gap

BioSymbols.AA_H — Constant.

Histidine

BioSymbols.AA_I — Constant.

Isoleucine

BioSymbols.AA_INVALID — Constant.

Invalid Amino Acid

BioSymbols.AA_J — Constant.

Leucine or Isoleucine

BioSymbols.AA_K — Constant.

Lysine

BioSymbols.AA_L — Constant.

Leucine

BioSymbols.AA_M — Constant.

Methionine

BioSymbols.AA_N — Constant.

Asparagine

BioSymbols.AA_O — Constant.

Pyrrolysine

BioSymbols.AA_P — Constant.

Proline

BioSymbols.AA_Q — Constant.

Glutamine

BioSymbols.AA_R — Constant.

Arginine

BioSymbols.AA_S — Constant.

Serine

BioSymbols.AA_T — Constant.

Threonine

BioSymbols.AA_Term — Constant.

Terminal

BioSymbols.AA_U — Constant.

Selenocysteine

BioSymbols.AA_V — Constant.

Valine

BioSymbols.AA_W — Constant.

Tryptophan

BioSymbols.AA_X — Constant.

Unspecified or Unknown Amino Acid

BioSymbols.AA_Y — Constant.

Tyrosine

BioSymbols.AA_Z — Constant.

Glutamine or Glutamic Acid

BioSymbols.ACGT — Constant.

ACGT

Unambiguous DNA.

Examples

julia> ACGT
(DNA_A, DNA_C, DNA_G, DNA_T)

BioSymbols.ACGTN — Constant.

ACGTN

Unambiguous DNA and DNA_N.

Examples

julia> ACGTN
(DNA_A, DNA_C, DNA_G, DNA_T, DNA_N)

BioSymbols.ACGU — Constant.

ACGU

Unambiguous RNA.

Examples

julia> ACGU
(RNA_A, RNA_C, RNA_G, RNA_U)

BioSymbols.ACGUN — Constant.

ACGUN

Unambiguous RNA and RNA_N.

Examples

julia> ACGUN
(RNA_A, RNA_C, RNA_G, RNA_U, RNA_N)

BioSymbols.AminoAcid — Type.

An amino acid type.

BioSymbols.BioSymbols

BioSymbols.DNA — Type.

A deoxyribonucleic acid type.

BioSymbols.DNA_A — Constant.

DNA Adenine

BioSymbols.DNA_B — Constant.

DNA Cytosine, Guanine or Thymine

BioSymbols.DNA_C — Constant.

DNA Cytosine

BioSymbols.DNA_D — Constant.

DNA Adenine, Guanine or Thymine

BioSymbols.DNA_G — Constant.

DNA Guanine

BioSymbols.DNA_Gap — Constant.

DNA Gap

BioSymbols.DNA_H — Constant.

DNA Adenine, Cytosine or Thymine

BioSymbols.DNA_K — Constant.

DNA Guanine or Thymine

BioSymbols.DNA_M — Constant.

DNA Adenine or Cytosine

BioSymbols.DNA_N — Constant.

DNA Adenine, Cytosine, Guanine or Thymine

BioSymbols.DNA_R — Constant.

DNA Adenine or Guanine

BioSymbols.DNA_S — Constant.

DNA Cytosine or Guanine

BioSymbols.DNA_T — Constant.

DNA Thymine

BioSymbols.DNA_V — Constant.

DNA Adenine, Cytosine or Guanine

BioSymbols.DNA_W — Constant.

DNA Adenine or Thymine

BioSymbols.DNA_Y — Constant.

DNA Cytosine or Thymine

BioSymbols.NucleicAcid — Type.

An abstract nucleic acid type.

BioSymbols.RNA — Type.

A ribonucleic acid type.

BioSymbols.RNA_A — Constant.

RNA Adenine

BioSymbols.RNA_B — Constant.

RNA Cytosine, Guanine or Uracil

BioSymbols.RNA_C — Constant.

RNA Cytosine

BioSymbols.RNA_D — Constant.

RNA Adenine, Guanine or Uracil

BioSymbols.RNA_G — Constant.

RNA Guanine

BioSymbols.RNA_Gap — Constant.

RNA Gap

BioSymbols.RNA_H — Constant.

RNA Adenine, Cytosine or Uracil

BioSymbols.RNA_K — Constant.

RNA Guanine or Uracil

BioSymbols.RNA_M — Constant.

RNA Adenine or Cytosine

BioSymbols.RNA_N — Constant.

RNA Adenine, Cytosine, Guanine or Uracil

BioSymbols.RNA_R — Constant.

RNA Adenine or Guanine

BioSymbols.RNA_S — Constant.

RNA Cytosine or Guanine

BioSymbols.RNA_U — Constant.

RNA Uracil

BioSymbols.RNA_V — Constant.

RNA Adenine, Cytosine or Guanine

BioSymbols.RNA_W — Constant.

RNA Adenine or Uracil

BioSymbols.RNA_Y — Constant.

RNA Cytosine or Uracil

BioSymbols.aa_to_char

BioSymbols.alphabet — Function.

alphabet(DNA)

Get all symbols of DNA in sorted order.

Examples

julia> alphabet(DNA)
(DNA_Gap, DNA_A, DNA_C, DNA_M, DNA_G, DNA_R, DNA_S, DNA_V, DNA_T, DNA_W, DNA_Y, DNA_H, DNA_K, DNA_D, DNA_B, DNA_N)

julia> issorted(alphabet(DNA))
true

alphabet(RNA)

Get all symbols of RNA in sorted order.

Examples

julia> alphabet(RNA)
(RNA_Gap, RNA_A, RNA_C, RNA_M, RNA_G, RNA_R, RNA_S, RNA_V, RNA_U, RNA_W, RNA_Y, RNA_H, RNA_K, RNA_D, RNA_B, RNA_N)

julia> issorted(alphabet(RNA))
true

alphabet(AminoAcid)

Get all symbols of AminoAcid in sorted order.

Examples

julia> alphabet(AminoAcid)
(AA_A, AA_R, AA_N, AA_D, AA_C, AA_Q, AA_E, AA_G, AA_H, AA_I, AA_L, AA_K, AA_M, AA_F, AA_P, AA_S, AA_T, AA_W, AA_Y, AA_V, AA_O, AA_U, AA_B, AA_J, AA_Z, AA_X, AA_Term, AA_Gap)

julia> issorted(alphabet(AminoAcid))
true

BioSymbols.char_to_aa

BioSymbols.char_to_dna

BioSymbols.char_to_rna

BioSymbols.compatbits — Function.

compatbits(nt::NucleicAcid)

Return the compatibility bits of nt as UInt8.

Examples

julia> compatbits(DNA_A)
0x01

julia> compatbits(DNA_C)
0x02

julia> compatbits(DNA_N)
0x0f

compatbits(aa::AminoAcid)

Return the compatibility bits of aa as UInt32.

Examples

julia> compatbits(AA_A)
0x00000001

julia> compatbits(AA_J)
0x00000600

BioSymbols.compatbits_aa

BioSymbols.complement — Function.

complement(nt::NucleicAcid)

Return the complementary nucleotide of nt.

This function returns the union of all possible complementary nucleotides.

Examples

julia> complement(DNA_A)
DNA_T

julia> complement(DNA_N)
DNA_N

julia> complement(RNA_U)
RNA_A

BioSymbols.dna_to_char

BioSymbols.eval

BioSymbols.gap — Function.

gap(DNA)

Return DNA_Gap.

gap(RNA)

Return RNA_Gap.

gap(AminoAcid)

Return AA_Gap.

BioSymbols.include

BioSymbols.isGC — Function.

isGC(nt::NucleicAcid)

Test if nt is surely either guanine or cytosine.

BioSymbols.isambiguous — Function.

isambiguous(nt::NucleicAcid)

Test if nt is an ambiguous nucleotide.

isambiguous(aa::AminoAcid)

Test if aa is an ambiguous amino acid.

BioSymbols.iscertain — Function.

iscertain(nt::NucleicAcid)

Test if nt is a non-ambiguous nucleotide e.g. ACGT.

iscertain(aa::AminoAcid)

Test if aa is a non-ambiguous amino acid.

BioSymbols.iscompatible — Function.

iscompatible(x::T, y::T) where T <: NucleicAcid

Test if x and y are compatible with each other (i.e. x and y can be the same symbol).

x and y must be the same type.

Examples

julia> iscompatible(DNA_A, DNA_A)
true

julia> iscompatible(DNA_C, DNA_N)  # DNA_N can be DNA_C
true

julia> iscompatible(DNA_C, DNA_R)  # DNA_R (A or G) cannot be DNA_C
false

iscompatible(x::AminoAcid, y::AminoAcid)

Test if x and y are compatible with each other.

Examples

julia> iscompatible(AA_A, AA_R)
false

julia> iscompatible(AA_A, AA_X)
true

BioSymbols.isgap — Function.

isgap(nt::NucleicAcid)

Test if nt is a gap.

isgap(aa::AminoAcid)

Test if aa is a gap.

BioSymbols.ispurine — Function.

ispurine(nt::NucleicAcid)

Test if nt is surely a purine.

BioSymbols.ispyrimidine — Function.

ispyrimidine(nt::NucleicAcid)

Test if nt is surely a pyrimidine.

BioSymbols.prefix

BioSymbols.rna_to_char

BioSymbols.threeletter_to_aa