Docstrings

Discretizers.AbstractDiscretizer

Discretizers.AbstractSampleMethod

Discretizers.CategoricalDiscretizer

Discretizers.DEFAULT_LIN_DISC_FORCE_OUTLIERS_TO_CLOSEST

Discretizers.DiscretizationAlgorithm

Discretizers.DiscretizeBayesianBlocks

Discretizers.DiscretizeMODL

Discretizers.DiscretizeMODL_Greedy

Discretizers.DiscretizeMODL_Optimal

Discretizers.DiscretizeMODL_PostGreedy

Discretizers.DiscretizeUniformCount

Discretizers.DiscretizeUniformWidth — Type.

DiscretizeUniformWidth(alg::Symbol) If nbins is a symbol, automatically determine the number of bins to use

Discretizers.Discretizers

Discretizers.HybridDiscretizer

Discretizers.LinearDiscretizer — Type.

LinearDiscretizer

A discretizer which encodes (typically) continuous values into discrete bins. A univariate domain is divided into discrete by edges. A value V will be encoded into bin B if V ∈ [Bₗ Bᵣ) (or V ∈ [Bₗ Bᵣ] if B is the rightmost bin)

If forceoutliersto_closest is set to true, then values outside of binedges will be shunted into the nearest bin. Otherwise an error is thrown.

TODO(tim): handle NA handle Nullable handle NaN ability to specify decoding behavior

Discretizers.MODL_value2_oneintval

Discretizers.PQ_methods_updata

Discretizers.SAMPLE_BIN_CENTER

Discretizers.SAMPLE_UNIFORM

Discretizers.SampleBinCenter

Discretizers.SampleUniform

Discretizers.bincenters

Discretizers.binedges

Discretizers.binwidth

Discretizers.binwidths

Discretizers.datalineardiscretizer — Function.

a hybrid discretizers that maps a special missing value indicator (typically Inf or NaN) to a discrete bin, but otherwise is a linear discretizer

Discretizers.decode — Function.

There are several methods for decoding a LinearDiscretizer. The default is SampleUniform, which uniformly samples from the bin. SampleBinCenter returns the bin's center value.

Discretizers.decoded_type

Discretizers.encode

Discretizers.encoded_type

Discretizers.eval

Discretizers.get_discretization_counts

Discretizers.get_histogram_plot_arrays

Discretizers.get_nbins

Discretizers.greedy_merge

Discretizers.greedy_merge_index

Discretizers.include

Discretizers.insert_and_dedup!

Discretizers.merge_adj_intval

Discretizers.methods_merge

Discretizers.methods_merge2_split

Discretizers.methods_merge_split

Discretizers.methods_split

Discretizers.nlabels

Discretizers.optimal_result

Discretizers.post_greedy_index

Discretizers.post_greedy_result

Discretizers.remove_methods_on_index

Discretizers.supports_decoding

Discretizers.supports_encoding

Discretizers.totalwidth

Discretizers.uncondi_greedy_merge_index