Docstrings

Sundials.@checkflag — Macro.

Insert a check that the given function call returns 0,
throw an error otherwise. Only apply directly to function calls.

Sundials.ANONYMOUS_1

Sundials.ANONYMOUS_2

Sundials.ANONYMOUS_3

Sundials.ANONYMOUS_4

Sundials.ANONYMOUS_5

Sundials.ARK324L2SA_DIRK_4_2_3

Sundials.ARK324L2SA_ERK_4_2_3

Sundials.ARK436L2SA_DIRK_6_3_4

Sundials.ARK436L2SA_ERK_6_3_4

Sundials.ARK548L2SA_DIRK_8_4_5

Sundials.ARK548L2SA_ERK_8_4_5

Sundials.ARKAdaptFn

Sundials.ARKBBDPrecGetNumGfnEvals

Sundials.ARKBBDPrecGetWorkSpace

Sundials.ARKBBDPrecInit

Sundials.ARKBBDPrecReInit

Sundials.ARKBandPrecGetNumRhsEvals

Sundials.ARKBandPrecGetWorkSpace

Sundials.ARKBandPrecInit

Sundials.ARKCommFn

Sundials.ARKDLS_ILL_INPUT

Sundials.ARKDLS_JACFUNC_RECVR

Sundials.ARKDLS_JACFUNC_UNRECVR

Sundials.ARKDLS_LMEM_NULL

Sundials.ARKDLS_MASSFUNC_RECVR

Sundials.ARKDLS_MASSFUNC_UNRECVR

Sundials.ARKDLS_MASSMEM_NULL

Sundials.ARKDLS_MEM_FAIL

Sundials.ARKDLS_MEM_NULL

Sundials.ARKDLS_SUCCESS

Sundials.ARKDLS_SUNMAT_FAIL

Sundials.ARKDlsGetLastFlag

Sundials.ARKDlsGetLastMassFlag

Sundials.ARKDlsGetMassWorkSpace

Sundials.ARKDlsGetNumJacEvals

Sundials.ARKDlsGetNumMassMult

Sundials.ARKDlsGetNumMassSetups

Sundials.ARKDlsGetNumMassSolves

Sundials.ARKDlsGetNumRhsEvals

Sundials.ARKDlsGetReturnFlagName

Sundials.ARKDlsGetWorkSpace

Sundials.ARKDlsJacFn

Sundials.ARKDlsMassFn

Sundials.ARKDlsSetJacFn

Sundials.ARKDlsSetLinearSolver

Sundials.ARKDlsSetMassFn

Sundials.ARKDlsSetMassLinearSolver

Sundials.ARKErrHandlerFn

Sundials.ARKEwtFn

Sundials.ARKExpStabFn

Sundials.ARKLocalFn

Sundials.ARKODE

Sundials.ARKODEIntegrator

Sundials.ARKODEMem

Sundials.ARKODEMemPtr

Sundials.ARKODEh — Type.

Handle for Sundials objects (CVODE, IDA, KIN).

Wraps the reference to the pointer to the Sundials object. Manages automatic destruction of the referenced objects when it is no longer in use.

Sundials.ARKPostProcessStepFn

Sundials.ARKRhsFn

Sundials.ARKRhsFn_wrapper

Sundials.ARKRootFn

Sundials.ARKRwtFn

Sundials.ARKSPILS_ILL_INPUT

Sundials.ARKSPILS_LMEM_NULL

Sundials.ARKSPILS_MASSMEM_NULL

Sundials.ARKSPILS_MEM_FAIL

Sundials.ARKSPILS_MEM_NULL

Sundials.ARKSPILS_MSBPRE

Sundials.ARKSPILS_PMEM_NULL

Sundials.ARKSPILS_SUCCESS

Sundials.ARKSPILS_SUNLS_FAIL

Sundials.ARKSpilsGetLastFlag

Sundials.ARKSpilsGetLastMassFlag

Sundials.ARKSpilsGetMassWorkSpace

Sundials.ARKSpilsGetNumConvFails

Sundials.ARKSpilsGetNumJTSetupEvals

Sundials.ARKSpilsGetNumJtimesEvals

Sundials.ARKSpilsGetNumLinIters

Sundials.ARKSpilsGetNumMassConvFails

Sundials.ARKSpilsGetNumMassIters

Sundials.ARKSpilsGetNumMassPrecEvals

Sundials.ARKSpilsGetNumMassPrecSolves

Sundials.ARKSpilsGetNumMtimesEvals

Sundials.ARKSpilsGetNumPrecEvals

Sundials.ARKSpilsGetNumPrecSolves

Sundials.ARKSpilsGetNumRhsEvals

Sundials.ARKSpilsGetReturnFlagName

Sundials.ARKSpilsGetWorkSpace

Sundials.ARKSpilsJacTimesSetupFn

Sundials.ARKSpilsJacTimesVecFn

Sundials.ARKSpilsMassPrecSetupFn

Sundials.ARKSpilsMassPrecSolveFn

Sundials.ARKSpilsMassTimesSetupFn

Sundials.ARKSpilsMassTimesVecFn

Sundials.ARKSpilsPrecSetupFn

Sundials.ARKSpilsPrecSolveFn

Sundials.ARKSpilsSetEpsLin

Sundials.ARKSpilsSetJacTimes

Sundials.ARKSpilsSetLinearSolver

Sundials.ARKSpilsSetMassEpsLin

Sundials.ARKSpilsSetMassLinearSolver

Sundials.ARKSpilsSetMassPreconditioner

Sundials.ARKSpilsSetMassTimes

Sundials.ARKSpilsSetPreconditioner

Sundials.ARKVecResizeFn

Sundials.ARK_BAD_DKY

Sundials.ARK_BAD_K

Sundials.ARK_BAD_T

Sundials.ARK_CONV_FAILURE

Sundials.ARK_ERR_FAILURE

Sundials.ARK_FIRST_RHSFUNC_ERR

Sundials.ARK_ILL_INPUT

Sundials.ARK_LFREE_FAIL

Sundials.ARK_LINIT_FAIL

Sundials.ARK_LSETUP_FAIL

Sundials.ARK_LSOLVE_FAIL

Sundials.ARK_MASSFREE_FAIL

Sundials.ARK_MASSINIT_FAIL

Sundials.ARK_MASSMULT_FAIL

Sundials.ARK_MASSSETUP_FAIL

Sundials.ARK_MASSSOLVE_FAIL

Sundials.ARK_MEM_FAIL

Sundials.ARK_MEM_NULL

Sundials.ARK_NORMAL

Sundials.ARK_NO_MALLOC

Sundials.ARK_ONE_STEP

Sundials.ARK_POSTPROCESS_FAIL

Sundials.ARK_REPTD_RHSFUNC_ERR

Sundials.ARK_RHSFUNC_FAIL

Sundials.ARK_ROOT_RETURN

Sundials.ARK_RTFUNC_FAIL

Sundials.ARK_SUCCESS

Sundials.ARK_S_MAX

Sundials.ARK_TOO_CLOSE

Sundials.ARK_TOO_MUCH_ACC

Sundials.ARK_TOO_MUCH_WORK

Sundials.ARK_TSTOP_RETURN

Sundials.ARK_UNREC_RHSFUNC_ERR

Sundials.ARK_WARNING

Sundials.ARKode

Sundials.ARKodeCreate

Sundials.ARKodeFree

Sundials.ARKodeGetActualInitStep

Sundials.ARKodeGetCurrentButcherTables

Sundials.ARKodeGetCurrentStep

Sundials.ARKodeGetCurrentTime

Sundials.ARKodeGetDky

Sundials.ARKodeGetErrWeights

Sundials.ARKodeGetEstLocalErrors

Sundials.ARKodeGetIntegratorStats

Sundials.ARKodeGetLastStep

Sundials.ARKodeGetNonlinSolvStats

Sundials.ARKodeGetNumAccSteps

Sundials.ARKodeGetNumErrTestFails

Sundials.ARKodeGetNumExpSteps

Sundials.ARKodeGetNumGEvals

Sundials.ARKodeGetNumLinSolvSetups

Sundials.ARKodeGetNumMassMultiplies

Sundials.ARKodeGetNumMassSolves

Sundials.ARKodeGetNumNonlinSolvConvFails

Sundials.ARKodeGetNumNonlinSolvIters

Sundials.ARKodeGetNumRhsEvals

Sundials.ARKodeGetNumStepAttempts

Sundials.ARKodeGetNumSteps

Sundials.ARKodeGetResWeights

Sundials.ARKodeGetReturnFlagName

Sundials.ARKodeGetRootInfo

Sundials.ARKodeGetTolScaleFactor

Sundials.ARKodeGetWorkSpace

Sundials.ARKodeInit

Sundials.ARKodeLoadButcherTable

Sundials.ARKodeReInit

Sundials.ARKodeResFtolerance

Sundials.ARKodeResStolerance

Sundials.ARKodeResVtolerance

Sundials.ARKodeResize

Sundials.ARKodeRootInit

Sundials.ARKodeSStolerances

Sundials.ARKodeSVtolerances

Sundials.ARKodeSetARKTableNum

Sundials.ARKodeSetARKTables

Sundials.ARKodeSetAdaptivityFn

Sundials.ARKodeSetAdaptivityMethod

Sundials.ARKodeSetCFLFraction

Sundials.ARKodeSetDefaults

Sundials.ARKodeSetDeltaGammaMax

Sundials.ARKodeSetDenseOrder

Sundials.ARKodeSetDiagnostics

Sundials.ARKodeSetERKTable

Sundials.ARKodeSetERKTableNum

Sundials.ARKodeSetErrFile

Sundials.ARKodeSetErrHandlerFn

Sundials.ARKodeSetErrorBias

Sundials.ARKodeSetExplicit

Sundials.ARKodeSetFixedPoint

Sundials.ARKodeSetFixedStep

Sundials.ARKodeSetFixedStepBounds

Sundials.ARKodeSetIRKTable

Sundials.ARKodeSetIRKTableNum

Sundials.ARKodeSetImEx

Sundials.ARKodeSetImplicit

Sundials.ARKodeSetInitStep

Sundials.ARKodeSetLinear

Sundials.ARKodeSetMaxCFailGrowth

Sundials.ARKodeSetMaxConvFails

Sundials.ARKodeSetMaxEFailGrowth

Sundials.ARKodeSetMaxErrTestFails

Sundials.ARKodeSetMaxFirstGrowth

Sundials.ARKodeSetMaxGrowth

Sundials.ARKodeSetMaxHnilWarns

Sundials.ARKodeSetMaxNonlinIters

Sundials.ARKodeSetMaxNumSteps

Sundials.ARKodeSetMaxStep

Sundials.ARKodeSetMaxStepsBetweenLSet

Sundials.ARKodeSetMinStep

Sundials.ARKodeSetNewton

Sundials.ARKodeSetNoInactiveRootWarn

Sundials.ARKodeSetNonlinCRDown

Sundials.ARKodeSetNonlinConvCoef

Sundials.ARKodeSetNonlinRDiv

Sundials.ARKodeSetNonlinear

Sundials.ARKodeSetOptimalParams

Sundials.ARKodeSetOrder

Sundials.ARKodeSetPostprocessStepFn

Sundials.ARKodeSetPredictorMethod

Sundials.ARKodeSetRootDirection

Sundials.ARKodeSetSafetyFactor

Sundials.ARKodeSetSmallNumEFails

Sundials.ARKodeSetStabilityFn

Sundials.ARKodeSetStopTime

Sundials.ARKodeSetUserData

Sundials.ARKodeWFtolerances

Sundials.ARKodeWriteButcher

Sundials.ARKodeWriteParameters

Sundials.ATimesFn

Sundials.AbstactFunJac

Sundials.AbstractAnalyticalProblem

Sundials.AbstractSundialsIntegrator

Sundials.AbstractSundialsObject — Type.

Base type for dummy placeholders that help to
providing typed pointers for Sundials objects
(KINSOL, CVODE, IDA).

See `Handle`.

Sundials.AffineDiffEqOperator

Sundials.AnalyticalProblem

Sundials.BIG_REAL

Sundials.BILLINGTON_3_3_2

Sundials.BOGACKI_SHAMPINE_4_2_3

Sundials.BVProblem

Sundials.Band

Sundials.BandMatrix

Sundials.CASH_5_2_4

Sundials.CASH_5_3_4

Sundials.CASH_KARP_6_4_5

Sundials.CLASSICAL_GS

Sundials.CSC_MAT

Sundials.CSR_MAT

Sundials.CVBBDPrecGetNumGfnEvals

Sundials.CVBBDPrecGetWorkSpace

Sundials.CVBBDPrecInit

Sundials.CVBBDPrecInitB

Sundials.CVBBDPrecReInit

Sundials.CVBBDPrecReInitB

Sundials.CVBandPrecGetNumRhsEvals

Sundials.CVBandPrecGetWorkSpace

Sundials.CVBandPrecInit

Sundials.CVBandPrecInitB

Sundials.CVCommFn

Sundials.CVCommFnB

Sundials.CVDIAG_ILL_INPUT

Sundials.CVDIAG_INV_FAIL

Sundials.CVDIAG_LMEM_NULL

Sundials.CVDIAG_MEM_FAIL

Sundials.CVDIAG_MEM_NULL

Sundials.CVDIAG_NO_ADJ

Sundials.CVDIAG_RHSFUNC_RECVR

Sundials.CVDIAG_RHSFUNC_UNRECVR

Sundials.CVDIAG_SUCCESS

Sundials.CVDLS_ILL_INPUT

Sundials.CVDLS_JACFUNC_RECVR

Sundials.CVDLS_JACFUNC_UNRECVR

Sundials.CVDLS_LMEMB_NULL

Sundials.CVDLS_LMEM_NULL

Sundials.CVDLS_MEM_FAIL

Sundials.CVDLS_MEM_NULL

Sundials.CVDLS_NO_ADJ

Sundials.CVDLS_SUCCESS

Sundials.CVDLS_SUNMAT_FAIL

Sundials.CVDiag

Sundials.CVDiagB

Sundials.CVDiagGetLastFlag

Sundials.CVDiagGetNumRhsEvals

Sundials.CVDiagGetReturnFlagName

Sundials.CVDiagGetWorkSpace

Sundials.CVDlsGetLastFlag

Sundials.CVDlsGetNumJacEvals

Sundials.CVDlsGetNumRhsEvals

Sundials.CVDlsGetReturnFlagName

Sundials.CVDlsGetWorkSpace

Sundials.CVDlsJacFn

Sundials.CVDlsJacFnB

Sundials.CVDlsJacFnBS

Sundials.CVDlsSetJacFn

Sundials.CVDlsSetJacFnB

Sundials.CVDlsSetJacFnBS

Sundials.CVDlsSetLinearSolver

Sundials.CVDlsSetLinearSolverB

Sundials.CVErrHandlerFn

Sundials.CVEwtFn

Sundials.CVLocalFn

Sundials.CVLocalFnB

Sundials.CVODEIntegrator

Sundials.CVODEMem

Sundials.CVODEMemPtr

Sundials.CVODE_Adams

Sundials.CVODE_BDF

Sundials.CVODEh — Type.

Handle for Sundials objects (CVODE, IDA, KIN).

Wraps the reference to the pointer to the Sundials object. Manages automatic destruction of the referenced objects when it is no longer in use.

Sundials.CVQuadRhsFn

Sundials.CVQuadRhsFnB

Sundials.CVQuadRhsFnBS

Sundials.CVQuadRhsFn_wrapper

Sundials.CVQuadSensRhsFn

Sundials.CVRhsFn

Sundials.CVRhsFnB

Sundials.CVRhsFnBS

Sundials.CVRhsFn_wrapper

Sundials.CVRootFn

Sundials.CVRootFn_wrapper

Sundials.CVSPILS_ILL_INPUT

Sundials.CVSPILS_LMEMB_NULL

Sundials.CVSPILS_LMEM_NULL

Sundials.CVSPILS_MEM_FAIL

Sundials.CVSPILS_MEM_NULL

Sundials.CVSPILS_MSBPRE

Sundials.CVSPILS_NO_ADJ

Sundials.CVSPILS_PMEM_NULL

Sundials.CVSPILS_SUCCESS

Sundials.CVSPILS_SUNLS_FAIL

Sundials.CVSensRhs1Fn

Sundials.CVSensRhsFn

Sundials.CVSpilsGetLastFlag

Sundials.CVSpilsGetNumConvFails

Sundials.CVSpilsGetNumJTSetupEvals

Sundials.CVSpilsGetNumJtimesEvals

Sundials.CVSpilsGetNumLinIters

Sundials.CVSpilsGetNumPrecEvals

Sundials.CVSpilsGetNumPrecSolves

Sundials.CVSpilsGetNumRhsEvals

Sundials.CVSpilsGetReturnFlagName

Sundials.CVSpilsGetWorkSpace

Sundials.CVSpilsJacTimesSetupFn

Sundials.CVSpilsJacTimesSetupFnB

Sundials.CVSpilsJacTimesSetupFnBS

Sundials.CVSpilsJacTimesVecFn

Sundials.CVSpilsJacTimesVecFnB

Sundials.CVSpilsJacTimesVecFnBS

Sundials.CVSpilsPrecSetupFn

Sundials.CVSpilsPrecSetupFnB

Sundials.CVSpilsPrecSetupFnBS

Sundials.CVSpilsPrecSolveFn

Sundials.CVSpilsPrecSolveFnB

Sundials.CVSpilsPrecSolveFnBS

Sundials.CVSpilsSetEpsLin

Sundials.CVSpilsSetEpsLinB

Sundials.CVSpilsSetJacTimes

Sundials.CVSpilsSetJacTimesB

Sundials.CVSpilsSetJacTimesBS

Sundials.CVSpilsSetLinearSolver

Sundials.CVSpilsSetLinearSolverB

Sundials.CVSpilsSetPreconditioner

Sundials.CVSpilsSetPreconditionerB

Sundials.CVSpilsSetPreconditionerBS

Sundials.CV_ADAMS

Sundials.CV_BAD_DKY

Sundials.CV_BAD_IS

Sundials.CV_BAD_K

Sundials.CV_BAD_T

Sundials.CV_BAD_TB0

Sundials.CV_BDF

Sundials.CV_CENTERED

Sundials.CV_CONV_FAILURE

Sundials.CV_ERR_FAILURE

Sundials.CV_FIRST_QRHSFUNC_ERR

Sundials.CV_FIRST_QSRHSFUNC_ERR

Sundials.CV_FIRST_RHSFUNC_ERR

Sundials.CV_FIRST_SRHSFUNC_ERR

Sundials.CV_FORWARD

Sundials.CV_FUNCTIONAL

Sundials.CV_FWD_FAIL

Sundials.CV_GETY_BADT

Sundials.CV_HERMITE

Sundials.CV_ILL_INPUT

Sundials.CV_LINIT_FAIL

Sundials.CV_LSETUP_FAIL

Sundials.CV_LSOLVE_FAIL

Sundials.CV_MEM_FAIL

Sundials.CV_MEM_NULL

Sundials.CV_NEWTON

Sundials.CV_NORMAL

Sundials.CV_NO_ADJ

Sundials.CV_NO_BCK

Sundials.CV_NO_FWD

Sundials.CV_NO_MALLOC

Sundials.CV_NO_QUAD

Sundials.CV_NO_QUADSENS

Sundials.CV_NO_SENS

Sundials.CV_ONE_STEP

Sundials.CV_POLYNOMIAL

Sundials.CV_QRHSFUNC_FAIL

Sundials.CV_QSRHSFUNC_FAIL

Sundials.CV_REIFWD_FAIL

Sundials.CV_REPTD_QRHSFUNC_ERR

Sundials.CV_REPTD_QSRHSFUNC_ERR

Sundials.CV_REPTD_RHSFUNC_ERR

Sundials.CV_REPTD_SRHSFUNC_ERR

Sundials.CV_RHSFUNC_FAIL

Sundials.CV_ROOT_RETURN

Sundials.CV_RTFUNC_FAIL

Sundials.CV_SIMULTANEOUS

Sundials.CV_SRHSFUNC_FAIL

Sundials.CV_STAGGERED

Sundials.CV_STAGGERED1

Sundials.CV_SUCCESS

Sundials.CV_TOO_CLOSE

Sundials.CV_TOO_MUCH_ACC

Sundials.CV_TOO_MUCH_WORK

Sundials.CV_TSTOP_RETURN

Sundials.CV_UNREC_QRHSFUNC_ERR

Sundials.CV_UNREC_QSRHSFUNC_ERR

Sundials.CV_UNREC_RHSFUNC_ERR

Sundials.CV_UNREC_SRHSFUNC_ERR

Sundials.CV_WARNING

Sundials.CVadjCheckPointRec

Sundials.CVode

Sundials.CVodeAdjFree

Sundials.CVodeAdjInit

Sundials.CVodeAdjReInit

Sundials.CVodeB

Sundials.CVodeCreate

Sundials.CVodeCreateB

Sundials.CVodeF

Sundials.CVodeFree

Sundials.CVodeGetActualInitStep

Sundials.CVodeGetAdjCVodeBmem

Sundials.CVodeGetAdjCheckPointsInfo

Sundials.CVodeGetAdjCurrentCheckPoint

Sundials.CVodeGetAdjDataPointHermite

Sundials.CVodeGetAdjDataPointPolynomial

Sundials.CVodeGetAdjY

Sundials.CVodeGetB

Sundials.CVodeGetCurrentOrder

Sundials.CVodeGetCurrentStep

Sundials.CVodeGetCurrentTime

Sundials.CVodeGetDky

Sundials.CVodeGetErrWeights

Sundials.CVodeGetEstLocalErrors

Sundials.CVodeGetIntegratorStats

Sundials.CVodeGetLastOrder

Sundials.CVodeGetLastStep

Sundials.CVodeGetNonlinSolvStats

Sundials.CVodeGetNumErrTestFails

Sundials.CVodeGetNumGEvals

Sundials.CVodeGetNumLinSolvSetups

Sundials.CVodeGetNumNonlinSolvConvFails

Sundials.CVodeGetNumNonlinSolvIters

Sundials.CVodeGetNumRhsEvals

Sundials.CVodeGetNumRhsEvalsSens

Sundials.CVodeGetNumStabLimOrderReds

Sundials.CVodeGetNumSteps

Sundials.CVodeGetQuad

Sundials.CVodeGetQuadB

Sundials.CVodeGetQuadDky

Sundials.CVodeGetQuadErrWeights

Sundials.CVodeGetQuadNumErrTestFails

Sundials.CVodeGetQuadNumRhsEvals

Sundials.CVodeGetQuadSens

Sundials.CVodeGetQuadSens1

Sundials.CVodeGetQuadSensDky

Sundials.CVodeGetQuadSensDky1

Sundials.CVodeGetQuadSensErrWeights

Sundials.CVodeGetQuadSensNumErrTestFails

Sundials.CVodeGetQuadSensNumRhsEvals

Sundials.CVodeGetQuadSensStats

Sundials.CVodeGetQuadStats

Sundials.CVodeGetReturnFlagName

Sundials.CVodeGetRootInfo

Sundials.CVodeGetSens

Sundials.CVodeGetSens1

Sundials.CVodeGetSensDky

Sundials.CVodeGetSensDky1

Sundials.CVodeGetSensErrWeights

Sundials.CVodeGetSensNonlinSolvStats

Sundials.CVodeGetSensNumErrTestFails

Sundials.CVodeGetSensNumLinSolvSetups

Sundials.CVodeGetSensNumNonlinSolvConvFails

Sundials.CVodeGetSensNumNonlinSolvIters

Sundials.CVodeGetSensNumRhsEvals

Sundials.CVodeGetSensStats

Sundials.CVodeGetStgrSensNumNonlinSolvConvFails

Sundials.CVodeGetStgrSensNumNonlinSolvIters

Sundials.CVodeGetTolScaleFactor

Sundials.CVodeGetWorkSpace

Sundials.CVodeInit

Sundials.CVodeInitB

Sundials.CVodeInitBS

Sundials.CVodeQuadFree

Sundials.CVodeQuadInit

Sundials.CVodeQuadInitB

Sundials.CVodeQuadInitBS

Sundials.CVodeQuadReInit

Sundials.CVodeQuadReInitB

Sundials.CVodeQuadSStolerances

Sundials.CVodeQuadSStolerancesB

Sundials.CVodeQuadSVtolerances

Sundials.CVodeQuadSVtolerancesB

Sundials.CVodeQuadSensEEtolerances

Sundials.CVodeQuadSensFree

Sundials.CVodeQuadSensInit

Sundials.CVodeQuadSensReInit

Sundials.CVodeQuadSensSStolerances

Sundials.CVodeQuadSensSVtolerances

Sundials.CVodeReInit

Sundials.CVodeReInitB

Sundials.CVodeRootInit

Sundials.CVodeSStolerances

Sundials.CVodeSStolerancesB

Sundials.CVodeSVtolerances

Sundials.CVodeSVtolerancesB

Sundials.CVodeSensEEtolerances

Sundials.CVodeSensFree

Sundials.CVodeSensInit

Sundials.CVodeSensInit1

Sundials.CVodeSensReInit

Sundials.CVodeSensSStolerances

Sundials.CVodeSensSVtolerances

Sundials.CVodeSensToggleOff

Sundials.CVodeSetAdjNoSensi

Sundials.CVodeSetErrFile

Sundials.CVodeSetErrHandlerFn

Sundials.CVodeSetInitStep

Sundials.CVodeSetInitStepB

Sundials.CVodeSetIterType

Sundials.CVodeSetIterTypeB

Sundials.CVodeSetMaxConvFails

Sundials.CVodeSetMaxErrTestFails

Sundials.CVodeSetMaxHnilWarns

Sundials.CVodeSetMaxNonlinIters

Sundials.CVodeSetMaxNumSteps

Sundials.CVodeSetMaxNumStepsB

Sundials.CVodeSetMaxOrd

Sundials.CVodeSetMaxOrdB

Sundials.CVodeSetMaxStep

Sundials.CVodeSetMaxStepB

Sundials.CVodeSetMinStep

Sundials.CVodeSetMinStepB

Sundials.CVodeSetNoInactiveRootWarn

Sundials.CVodeSetNonlinConvCoef

Sundials.CVodeSetQuadErrCon

Sundials.CVodeSetQuadErrConB

Sundials.CVodeSetQuadSensErrCon

Sundials.CVodeSetRootDirection

Sundials.CVodeSetSensDQMethod

Sundials.CVodeSetSensErrCon

Sundials.CVodeSetSensMaxNonlinIters

Sundials.CVodeSetSensParams

Sundials.CVodeSetStabLimDet

Sundials.CVodeSetStabLimDetB

Sundials.CVodeSetStopTime

Sundials.CVodeSetUserData

Sundials.CVodeSetUserDataB

Sundials.CVodeWFtolerances

Sundials.CallbackSet

Sundials.ContinuousCallback

Sundials.DAEFunction

Sundials.DAEProblem

Sundials.DAESolution

Sundials.DBL_EPSILON

Sundials.DBL_MAX

Sundials.DBL_MIN

Sundials.DDEFunction

Sundials.DDEProblem

Sundials.DEDataArray

Sundials.DEDataMatrix

Sundials.DEDataVector

Sundials.DEFAULT_ARK_ETABLE_3

Sundials.DEFAULT_ARK_ETABLE_4

Sundials.DEFAULT_ARK_ETABLE_5

Sundials.DEFAULT_ARK_ITABLE_3

Sundials.DEFAULT_ARK_ITABLE_4

Sundials.DEFAULT_ARK_ITABLE_5

Sundials.DEFAULT_DIRK_2

Sundials.DEFAULT_DIRK_3

Sundials.DEFAULT_DIRK_4

Sundials.DEFAULT_DIRK_5

Sundials.DEFAULT_ERK_2

Sundials.DEFAULT_ERK_3

Sundials.DEFAULT_ERK_4

Sundials.DEFAULT_ERK_5

Sundials.DEFAULT_ERK_6

Sundials.DEFAULT_ERK_8

Sundials.DEFAULT_LINSOLVE

Sundials.DEOptions

Sundials.DESolution

Sundials.DORMAND_PRINCE_7_4_5

Sundials.Dense

Sundials.DenseMatrix

Sundials.DiffEqBase

Sundials.DiscreteCallback

Sundials.DiscreteFunction

Sundials.DiscreteProblem

Sundials.DlsMat

Sundials.DynamicalODEFunction

Sundials.DynamicalODEProblem

Sundials.Explicit

Sundials.FCMIX_ARKODE

Sundials.FCMIX_CVODE

Sundials.FCMIX_IDA

Sundials.FCMIX_KINSOL

Sundials.FEHLBERG_13_7_8

Sundials.FEHLBERG_6_4_5

Sundials.FILE

Sundials.FunJac

Sundials.HEUN_EULER_2_1_2

Sundials.Handle — Type.

Handle for Sundials objects (CVODE, IDA, KIN).

Wraps the reference to the pointer to the Sundials object. Manages automatic destruction of the referenced objects when it is no longer in use.

Sundials.IDA

Sundials.IDAAdjFree

Sundials.IDAAdjInit

Sundials.IDAAdjReInit

Sundials.IDAAdjSetNoSensi

Sundials.IDABBDCommFn

Sundials.IDABBDCommFnB

Sundials.IDABBDLocalFn

Sundials.IDABBDLocalFnB

Sundials.IDABBDPrecGetNumGfnEvals

Sundials.IDABBDPrecGetWorkSpace

Sundials.IDABBDPrecInit

Sundials.IDABBDPrecInitB

Sundials.IDABBDPrecReInit

Sundials.IDABBDPrecReInitB

Sundials.IDACalcIC

Sundials.IDACalcICB

Sundials.IDACalcICBS

Sundials.IDACreate

Sundials.IDACreateB

Sundials.IDADLS_ILL_INPUT

Sundials.IDADLS_JACFUNC_RECVR

Sundials.IDADLS_JACFUNC_UNRECVR

Sundials.IDADLS_LMEMB_NULL

Sundials.IDADLS_LMEM_NULL

Sundials.IDADLS_MEM_FAIL

Sundials.IDADLS_MEM_NULL

Sundials.IDADLS_NO_ADJ

Sundials.IDADLS_SUCCESS

Sundials.IDADLS_SUNMAT_FAIL

Sundials.IDADlsGetLastFlag

Sundials.IDADlsGetNumJacEvals

Sundials.IDADlsGetNumResEvals

Sundials.IDADlsGetReturnFlagName

Sundials.IDADlsGetWorkSpace

Sundials.IDADlsJacFn

Sundials.IDADlsJacFnB

Sundials.IDADlsJacFnBS

Sundials.IDADlsSetJacFn

Sundials.IDADlsSetJacFnB

Sundials.IDADlsSetJacFnBS

Sundials.IDADlsSetLinearSolver

Sundials.IDADlsSetLinearSolverB

Sundials.IDAErrHandlerFn

Sundials.IDAEwtFn

Sundials.IDAFree

Sundials.IDAGetActualInitStep

Sundials.IDAGetAdjCheckPointsInfo

Sundials.IDAGetAdjCurrentCheckPoint

Sundials.IDAGetAdjDataPointHermite

Sundials.IDAGetAdjDataPointPolynomial

Sundials.IDAGetAdjIDABmem

Sundials.IDAGetAdjY

Sundials.IDAGetB

Sundials.IDAGetConsistentIC

Sundials.IDAGetConsistentICB

Sundials.IDAGetCurrentOrder

Sundials.IDAGetCurrentStep

Sundials.IDAGetCurrentTime

Sundials.IDAGetDky

Sundials.IDAGetErrWeights

Sundials.IDAGetEstLocalErrors

Sundials.IDAGetIntegratorStats

Sundials.IDAGetLastOrder

Sundials.IDAGetLastStep

Sundials.IDAGetNonlinSolvStats

Sundials.IDAGetNumBacktrackOps

Sundials.IDAGetNumErrTestFails

Sundials.IDAGetNumGEvals

Sundials.IDAGetNumLinSolvSetups

Sundials.IDAGetNumNonlinSolvConvFails

Sundials.IDAGetNumNonlinSolvIters

Sundials.IDAGetNumResEvals

Sundials.IDAGetNumResEvalsSens

Sundials.IDAGetNumSteps

Sundials.IDAGetQuad

Sundials.IDAGetQuadB

Sundials.IDAGetQuadDky

Sundials.IDAGetQuadErrWeights

Sundials.IDAGetQuadNumErrTestFails

Sundials.IDAGetQuadNumRhsEvals

Sundials.IDAGetQuadSens

Sundials.IDAGetQuadSens1

Sundials.IDAGetQuadSensDky

Sundials.IDAGetQuadSensDky1

Sundials.IDAGetQuadSensErrWeights

Sundials.IDAGetQuadSensNumErrTestFails

Sundials.IDAGetQuadSensNumRhsEvals

Sundials.IDAGetQuadSensStats

Sundials.IDAGetQuadStats

Sundials.IDAGetReturnFlagName

Sundials.IDAGetRootInfo

Sundials.IDAGetSens

Sundials.IDAGetSens1

Sundials.IDAGetSensConsistentIC

Sundials.IDAGetSensDky

Sundials.IDAGetSensDky1

Sundials.IDAGetSensErrWeights

Sundials.IDAGetSensNonlinSolvStats

Sundials.IDAGetSensNumErrTestFails

Sundials.IDAGetSensNumLinSolvSetups

Sundials.IDAGetSensNumNonlinSolvConvFails

Sundials.IDAGetSensNumNonlinSolvIters

Sundials.IDAGetSensNumResEvals

Sundials.IDAGetSensStats

Sundials.IDAGetTolScaleFactor

Sundials.IDAGetWorkSpace

Sundials.IDAInit

Sundials.IDAInitB

Sundials.IDAInitBS

Sundials.IDAIntegrator

Sundials.IDAMem

Sundials.IDAMemPtr

Sundials.IDAQuadFree

Sundials.IDAQuadInit

Sundials.IDAQuadInitB

Sundials.IDAQuadInitBS

Sundials.IDAQuadReInit

Sundials.IDAQuadReInitB

Sundials.IDAQuadRhsFn

Sundials.IDAQuadRhsFnB

Sundials.IDAQuadRhsFnBS

Sundials.IDAQuadSStolerances

Sundials.IDAQuadSStolerancesB

Sundials.IDAQuadSVtolerances

Sundials.IDAQuadSVtolerancesB

Sundials.IDAQuadSensEEtolerances

Sundials.IDAQuadSensFree

Sundials.IDAQuadSensInit

Sundials.IDAQuadSensReInit

Sundials.IDAQuadSensRhsFn

Sundials.IDAQuadSensSStolerances

Sundials.IDAQuadSensSVtolerances

Sundials.IDAReInit

Sundials.IDAReInitB

Sundials.IDAResFn

Sundials.IDAResFnB

Sundials.IDAResFnBS

Sundials.IDAResFn_wrapper

Sundials.IDARootFn

Sundials.IDARootFn_wrapper

Sundials.IDARootInit

Sundials.IDASPILS_ILL_INPUT

Sundials.IDASPILS_LMEMB_NULL

Sundials.IDASPILS_LMEM_NULL

Sundials.IDASPILS_MEM_FAIL

Sundials.IDASPILS_MEM_NULL

Sundials.IDASPILS_NO_ADJ

Sundials.IDASPILS_PMEM_NULL

Sundials.IDASPILS_SUCCESS

Sundials.IDASPILS_SUNLS_FAIL

Sundials.IDASStolerances

Sundials.IDASStolerancesB

Sundials.IDASVtolerances

Sundials.IDASVtolerancesB

Sundials.IDASensEEtolerances

Sundials.IDASensFree

Sundials.IDASensInit

Sundials.IDASensReInit

Sundials.IDASensResFn

Sundials.IDASensSStolerances

Sundials.IDASensSVtolerances

Sundials.IDASensToggleOff

Sundials.IDASetConstraints

Sundials.IDASetConstraintsB

Sundials.IDASetErrFile

Sundials.IDASetErrHandlerFn

Sundials.IDASetId

Sundials.IDASetIdB

Sundials.IDASetInitStep

Sundials.IDASetInitStepB

Sundials.IDASetLineSearchOffIC

Sundials.IDASetMaxBacksIC

Sundials.IDASetMaxConvFails

Sundials.IDASetMaxErrTestFails

Sundials.IDASetMaxNonlinIters

Sundials.IDASetMaxNumItersIC

Sundials.IDASetMaxNumJacsIC

Sundials.IDASetMaxNumSteps

Sundials.IDASetMaxNumStepsB

Sundials.IDASetMaxNumStepsIC

Sundials.IDASetMaxOrd

Sundials.IDASetMaxOrdB

Sundials.IDASetMaxStep

Sundials.IDASetMaxStepB

Sundials.IDASetNoInactiveRootWarn

Sundials.IDASetNonlinConvCoef

Sundials.IDASetNonlinConvCoefIC

Sundials.IDASetQuadErrCon

Sundials.IDASetQuadErrConB

Sundials.IDASetQuadSensErrCon

Sundials.IDASetRootDirection

Sundials.IDASetSensDQMethod

Sundials.IDASetSensErrCon

Sundials.IDASetSensMaxNonlinIters

Sundials.IDASetSensParams

Sundials.IDASetStepToleranceIC

Sundials.IDASetStopTime

Sundials.IDASetSuppressAlg

Sundials.IDASetSuppressAlgB

Sundials.IDASetUserData

Sundials.IDASetUserDataB

Sundials.IDASolve

Sundials.IDASolveB

Sundials.IDASolveF

Sundials.IDASpilsGetLastFlag

Sundials.IDASpilsGetNumConvFails

Sundials.IDASpilsGetNumJtimesEvals

Sundials.IDASpilsGetNumLinIters

Sundials.IDASpilsGetNumPrecEvals

Sundials.IDASpilsGetNumPrecSolves

Sundials.IDASpilsGetNumResEvals

Sundials.IDASpilsGetReturnFlagName

Sundials.IDASpilsGetWorkSpace

Sundials.IDASpilsJacTimesSetupFn

Sundials.IDASpilsJacTimesSetupFnB

Sundials.IDASpilsJacTimesSetupFnBS

Sundials.IDASpilsJacTimesVecFn

Sundials.IDASpilsJacTimesVecFnB

Sundials.IDASpilsJacTimesVecFnBS

Sundials.IDASpilsPrecSetupFn

Sundials.IDASpilsPrecSetupFnB

Sundials.IDASpilsPrecSetupFnBS

Sundials.IDASpilsPrecSolveFn

Sundials.IDASpilsPrecSolveFnB

Sundials.IDASpilsPrecSolveFnBS

Sundials.IDASpilsSetEpsLin

Sundials.IDASpilsSetEpsLinB

Sundials.IDASpilsSetIncrementFactor

Sundials.IDASpilsSetIncrementFactorB

Sundials.IDASpilsSetJacTimes

Sundials.IDASpilsSetJacTimesB

Sundials.IDASpilsSetJacTimesBS

Sundials.IDASpilsSetLinearSolver

Sundials.IDASpilsSetLinearSolverB

Sundials.IDASpilsSetPreconditioner

Sundials.IDASpilsSetPreconditionerB

Sundials.IDASpilsSetPreconditionerBS

Sundials.IDAWFtolerances

Sundials.IDA_BAD_DKY

Sundials.IDA_BAD_EWT

Sundials.IDA_BAD_IS

Sundials.IDA_BAD_K

Sundials.IDA_BAD_T

Sundials.IDA_BAD_TB0

Sundials.IDA_CENTERED

Sundials.IDA_CONSTR_FAIL

Sundials.IDA_CONV_FAIL

Sundials.IDA_ERR_FAIL

Sundials.IDA_FIRST_QRHS_ERR

Sundials.IDA_FIRST_QSRHS_ERR

Sundials.IDA_FIRST_RES_FAIL

Sundials.IDA_FORWARD

Sundials.IDA_FWD_FAIL

Sundials.IDA_GETY_BADT

Sundials.IDA_HERMITE

Sundials.IDA_ILL_INPUT

Sundials.IDA_LINESEARCH_FAIL

Sundials.IDA_LINIT_FAIL

Sundials.IDA_LSETUP_FAIL

Sundials.IDA_LSOLVE_FAIL

Sundials.IDA_MEM_FAIL

Sundials.IDA_MEM_NULL

Sundials.IDA_NORMAL

Sundials.IDA_NO_ADJ

Sundials.IDA_NO_BCK

Sundials.IDA_NO_FWD

Sundials.IDA_NO_MALLOC

Sundials.IDA_NO_QUAD

Sundials.IDA_NO_QUADSENS

Sundials.IDA_NO_RECOVERY

Sundials.IDA_NO_SENS

Sundials.IDA_ONE_STEP

Sundials.IDA_POLYNOMIAL

Sundials.IDA_QRHS_FAIL

Sundials.IDA_QSRHS_FAIL

Sundials.IDA_REIFWD_FAIL

Sundials.IDA_REP_QRHS_ERR

Sundials.IDA_REP_QSRHS_ERR

Sundials.IDA_REP_RES_ERR

Sundials.IDA_REP_SRES_ERR

Sundials.IDA_RES_FAIL

Sundials.IDA_ROOT_RETURN

Sundials.IDA_RTFUNC_FAIL

Sundials.IDA_SIMULTANEOUS

Sundials.IDA_SRES_FAIL

Sundials.IDA_STAGGERED

Sundials.IDA_SUCCESS

Sundials.IDA_TOO_MUCH_ACC

Sundials.IDA_TOO_MUCH_WORK

Sundials.IDA_TSTOP_RETURN

Sundials.IDA_UNRECOGNISED_ERROR

Sundials.IDA_UNRECONGISED_ERROR

Sundials.IDA_WARNING

Sundials.IDA_YA_YDP_INIT

Sundials.IDA_Y_INIT

Sundials.IDAadjCheckPointRec

Sundials.IDAh — Type.

Handle for Sundials objects (CVODE, IDA, KIN).

Wraps the reference to the pointer to the Sundials object. Manages automatic destruction of the referenced objects when it is no longer in use.

Sundials.Implicit

Sundials.KINBBDCommFn

Sundials.KINBBDLocalFn

Sundials.KINBBDPRE_FUNC_UNRECVR

Sundials.KINBBDPRE_PDATA_NULL

Sundials.KINBBDPRE_SUCCESS

Sundials.KINBBDPrecGetNumGfnEvals

Sundials.KINBBDPrecGetWorkSpace

Sundials.KINBBDPrecInit

Sundials.KINCreate

Sundials.KINDLS_ILL_INPUT

Sundials.KINDLS_JACFUNC_ERR

Sundials.KINDLS_LMEM_NULL

Sundials.KINDLS_MEM_FAIL

Sundials.KINDLS_MEM_NULL

Sundials.KINDLS_SUCCESS

Sundials.KINDLS_SUNMAT_FAIL

Sundials.KINDlsGetLastFlag

Sundials.KINDlsGetNumFuncEvals

Sundials.KINDlsGetNumJacEvals

Sundials.KINDlsGetReturnFlagName

Sundials.KINDlsGetWorkSpace

Sundials.KINDlsJacFn

Sundials.KINDlsSetJacFn

Sundials.KINDlsSetLinearSolver

Sundials.KINErrHandlerFn

Sundials.KINFree

Sundials.KINGetFuncNorm

Sundials.KINGetNumBacktrackOps

Sundials.KINGetNumBetaCondFails

Sundials.KINGetNumFuncEvals

Sundials.KINGetNumNonlinSolvIters

Sundials.KINGetReturnFlagName

Sundials.KINGetStepLength

Sundials.KINGetWorkSpace

Sundials.KINInfoHandlerFn

Sundials.KINInit

Sundials.KINMem

Sundials.KINMemPtr

Sundials.KINSPILS_ILL_INPUT

Sundials.KINSPILS_LMEM_NULL

Sundials.KINSPILS_MEM_FAIL

Sundials.KINSPILS_MEM_NULL

Sundials.KINSPILS_PMEM_NULL

Sundials.KINSPILS_SUCCESS

Sundials.KINSPILS_SUNLS_FAIL

Sundials.KINSetConstraints

Sundials.KINSetErrFile

Sundials.KINSetErrHandlerFn

Sundials.KINSetEtaConstValue

Sundials.KINSetEtaForm

Sundials.KINSetEtaParams

Sundials.KINSetFuncNormTol

Sundials.KINSetInfoFile

Sundials.KINSetInfoHandlerFn

Sundials.KINSetMAA

Sundials.KINSetMaxBetaFails

Sundials.KINSetMaxNewtonStep

Sundials.KINSetMaxSetupCalls

Sundials.KINSetMaxSubSetupCalls

Sundials.KINSetNoInitSetup

Sundials.KINSetNoMinEps

Sundials.KINSetNoResMon

Sundials.KINSetNumMaxIters

Sundials.KINSetPrintLevel

Sundials.KINSetRelErrFunc

Sundials.KINSetResMonConstValue

Sundials.KINSetResMonParams

Sundials.KINSetScaledStepTol

Sundials.KINSetSysFunc

Sundials.KINSetUserData

Sundials.KINSol

Sundials.KINSpilsGetLastFlag

Sundials.KINSpilsGetNumConvFails

Sundials.KINSpilsGetNumFuncEvals

Sundials.KINSpilsGetNumJtimesEvals

Sundials.KINSpilsGetNumLinIters

Sundials.KINSpilsGetNumPrecEvals

Sundials.KINSpilsGetNumPrecSolves

Sundials.KINSpilsGetReturnFlagName

Sundials.KINSpilsGetWorkSpace

Sundials.KINSpilsJacTimesVecFn

Sundials.KINSpilsPrecSetupFn

Sundials.KINSpilsPrecSolveFn

Sundials.KINSpilsSetJacTimesVecFn

Sundials.KINSpilsSetLinearSolver

Sundials.KINSpilsSetPreconditioner

Sundials.KINSysFn

Sundials.KINSysFn_wrapper

Sundials.KIN_ETACHOICE1

Sundials.KIN_ETACHOICE2

Sundials.KIN_ETACONSTANT

Sundials.KIN_FIRST_SYSFUNC_ERR

Sundials.KIN_FP

Sundials.KIN_ILL_INPUT

Sundials.KIN_INITIAL_GUESS_OK

Sundials.KIN_LINESEARCH

Sundials.KIN_LINESEARCH_BCFAIL

Sundials.KIN_LINESEARCH_NONCONV

Sundials.KIN_LINIT_FAIL

Sundials.KIN_LINSOLV_NO_RECOVERY

Sundials.KIN_LSETUP_FAIL

Sundials.KIN_LSOLVE_FAIL

Sundials.KIN_MAXITER_REACHED

Sundials.KIN_MEM_FAIL

Sundials.KIN_MEM_NULL

Sundials.KIN_MXNEWT_5X_EXCEEDED

Sundials.KIN_NONE

Sundials.KIN_NO_MALLOC

Sundials.KIN_PICARD

Sundials.KIN_REPTD_SYSFUNC_ERR

Sundials.KIN_STEP_LT_STPTOL

Sundials.KIN_SUCCESS

Sundials.KIN_SYSFUNC_FAIL

Sundials.KIN_WARNING

Sundials.KINh — Type.

Handle for Sundials objects (CVODE, IDA, KIN).

Wraps the reference to the pointer to the Sundials object. Manages automatic destruction of the referenced objects when it is no longer in use.

Sundials.KLU

Sundials.KVAERNO_4_2_3

Sundials.KVAERNO_5_3_4

Sundials.KVAERNO_7_4_5

Sundials.LinSolHandle

Sundials.LinSolveFactorize

Sundials.MAX_DIRK_NUM

Sundials.MAX_ERK_NUM

Sundials.MIN_DIRK_NUM

Sundials.MIN_ERK_NUM

Sundials.MODIFIED_GS

Sundials.MatrixHandle

Sundials.MonteCarloProblem

Sundials.MonteCarloSolution

Sundials.MonteCarloSummary

Sundials.MonteCarloTestSolution

Sundials.NVector — Type.

Wrapper for Sundials N_Vector that uses Julia Vector{realtype} as the data container.

Implements DenseVector interface and manages automatic destruction of the referenced N_Vector when it is no longer in use.

Sundials.N_VAbs

Sundials.N_VAbs_Serial

Sundials.N_VAddConst

Sundials.N_VAddConst_Serial

Sundials.N_VClone

Sundials.N_VCloneEmpty

Sundials.N_VCloneEmptyVectorArray

Sundials.N_VCloneEmpty_Serial

Sundials.N_VCloneVectorArray

Sundials.N_VCloneVectorArrayEmpty_Serial

Sundials.N_VCloneVectorArray_Serial

Sundials.N_VClone_Serial

Sundials.N_VCompare

Sundials.N_VCompare_Serial

Sundials.N_VConst

Sundials.N_VConst_Serial

Sundials.N_VConstrMask

Sundials.N_VConstrMask_Serial

Sundials.N_VDestroy

Sundials.N_VDestroyVectorArray

Sundials.N_VDestroyVectorArray_Serial

Sundials.N_VDestroy_Serial

Sundials.N_VDiv

Sundials.N_VDiv_Serial

Sundials.N_VDotProd

Sundials.N_VDotProd_Serial

Sundials.N_VGetArrayPointer

Sundials.N_VGetArrayPointer_Serial

Sundials.N_VGetLength_Serial

Sundials.N_VGetVectorID

Sundials.N_VGetVectorID_Serial

Sundials.N_VInv

Sundials.N_VInvTest

Sundials.N_VInvTest_Serial

Sundials.N_VInv_Serial

Sundials.N_VL1Norm

Sundials.N_VL1Norm_Serial

Sundials.N_VLinearSum

Sundials.N_VLinearSum_Serial

Sundials.N_VMake_Serial

Sundials.N_VMaxNorm

Sundials.N_VMaxNorm_Serial

Sundials.N_VMin

Sundials.N_VMinQuotient

Sundials.N_VMinQuotient_Serial

Sundials.N_VMin_Serial

Sundials.N_VNewEmpty_Serial

Sundials.N_VNew_Serial

Sundials.N_VPrintFile_Serial

Sundials.N_VPrint_Serial

Sundials.N_VProd

Sundials.N_VProd_Serial

Sundials.N_VScale

Sundials.N_VScale_Serial

Sundials.N_VSetArrayPointer

Sundials.N_VSetArrayPointer_Serial

Sundials.N_VSpace

Sundials.N_VSpace_Serial

Sundials.N_VWL2Norm

Sundials.N_VWL2Norm_Serial

Sundials.N_VWrmsNorm

Sundials.N_VWrmsNormMask

Sundials.N_VWrmsNormMask_Serial

Sundials.N_VWrmsNorm_Serial

Sundials.N_Vector

Sundials.N_VectorContent_Serial

Sundials.N_Vector_ID

Sundials.N_Vector_Ops

Sundials.N_Vector_S

Sundials.NoiseProblem

Sundials.ODEFunction

Sundials.ODEProblem

Sundials.ODESolution

Sundials.PCG

Sundials.PCG_ATIMES_FAIL_REC

Sundials.PCG_ATIMES_FAIL_UNREC

Sundials.PCG_CONV_FAIL

Sundials.PCG_MEM_NULL

Sundials.PCG_PSET_FAIL_REC

Sundials.PCG_PSET_FAIL_UNREC

Sundials.PCG_PSOLVE_FAIL_REC

Sundials.PCG_PSOLVE_FAIL_UNREC

Sundials.PCG_RES_REDUCED

Sundials.PCG_SUCCESS

Sundials.PREC_BOTH

Sundials.PREC_LEFT

Sundials.PREC_NONE

Sundials.PREC_RIGHT

Sundials.PSetupFn

Sundials.PSolveFn

Sundials.PTFQMR

Sundials.PcgMem

Sundials.PcgMemRec

Sundials.RODEFunction

Sundials.RODEProblem

Sundials.RODESolution

Sundials.SAYFY_ABURUB_6_3_4

Sundials.SDEFunction

Sundials.SDEProblem

Sundials.SDIRK_2_1_2

Sundials.SDIRK_5_3_4

Sundials.SMALL_REAL

Sundials.SPBCGS

Sundials.SPBCG_ATIMES_FAIL_REC

Sundials.SPBCG_ATIMES_FAIL_UNREC

Sundials.SPBCG_CONV_FAIL

Sundials.SPBCG_MEM_NULL

Sundials.SPBCG_PSET_FAIL_REC

Sundials.SPBCG_PSET_FAIL_UNREC

Sundials.SPBCG_PSOLVE_FAIL_REC

Sundials.SPBCG_PSOLVE_FAIL_UNREC

Sundials.SPBCG_RES_REDUCED

Sundials.SPBCG_SUCCESS

Sundials.SPFGMR

Sundials.SPFGMR_ATIMES_FAIL_REC

Sundials.SPFGMR_ATIMES_FAIL_UNREC

Sundials.SPFGMR_CONV_FAIL

Sundials.SPFGMR_GS_FAIL

Sundials.SPFGMR_MEM_NULL

Sundials.SPFGMR_PSET_FAIL_REC

Sundials.SPFGMR_PSET_FAIL_UNREC

Sundials.SPFGMR_PSOLVE_FAIL_REC

Sundials.SPFGMR_PSOLVE_FAIL_UNREC

Sundials.SPFGMR_QRFACT_FAIL

Sundials.SPFGMR_QRSOL_FAIL

Sundials.SPFGMR_RES_REDUCED

Sundials.SPFGMR_SUCCESS

Sundials.SPGMR

Sundials.SPGMR_ATIMES_FAIL_REC

Sundials.SPGMR_ATIMES_FAIL_UNREC

Sundials.SPGMR_CONV_FAIL

Sundials.SPGMR_GS_FAIL

Sundials.SPGMR_MEM_NULL

Sundials.SPGMR_PSET_FAIL_REC

Sundials.SPGMR_PSET_FAIL_UNREC

Sundials.SPGMR_PSOLVE_FAIL_REC

Sundials.SPGMR_PSOLVE_FAIL_UNREC

Sundials.SPGMR_QRFACT_FAIL

Sundials.SPGMR_QRSOL_FAIL

Sundials.SPGMR_RES_REDUCED

Sundials.SPGMR_SUCCESS

Sundials.SPTFQMR_ATIMES_FAIL_REC

Sundials.SPTFQMR_ATIMES_FAIL_UNREC

Sundials.SPTFQMR_CONV_FAIL

Sundials.SPTFQMR_MEM_NULL

Sundials.SPTFQMR_PSET_FAIL_REC

Sundials.SPTFQMR_PSET_FAIL_UNREC

Sundials.SPTFQMR_PSOLVE_FAIL_REC

Sundials.SPTFQMR_PSOLVE_FAIL_UNREC

Sundials.SPTFQMR_RES_REDUCED

Sundials.SPTFQMR_SUCCESS

Sundials.SUNBandLinearSolver

Sundials.SUNBandMatrix

Sundials.SUNBandMatrix_Cols

Sundials.SUNBandMatrix_Column

Sundials.SUNBandMatrix_Columns

Sundials.SUNBandMatrix_Data

Sundials.SUNBandMatrix_LDim

Sundials.SUNBandMatrix_LowerBandwidth

Sundials.SUNBandMatrix_Print

Sundials.SUNBandMatrix_Rows

Sundials.SUNBandMatrix_StoredUpperBandwidth

Sundials.SUNBandMatrix_UpperBandwidth

Sundials.SUNDIALSGetVersion

Sundials.SUNDIALSGetVersionNumber

Sundials.SUNDIALS_BAND

Sundials.SUNDIALS_DENSE

Sundials.SUNDIALS_DOUBLE_PRECISION

Sundials.SUNDIALS_INT64_T

Sundials.SUNDIALS_NVEC_CUDA

Sundials.SUNDIALS_NVEC_CUSTOM

Sundials.SUNDIALS_NVEC_OPENMP

Sundials.SUNDIALS_NVEC_PARALLEL

Sundials.SUNDIALS_NVEC_PARHYP

Sundials.SUNDIALS_NVEC_PETSC

Sundials.SUNDIALS_NVEC_PTHREADS

Sundials.SUNDIALS_NVEC_RAJA

Sundials.SUNDIALS_NVEC_SERIAL

Sundials.SUNDIALS_VERSION

Sundials.SUNDIALS_VERSION_LABEL

Sundials.SUNDIALS_VERSION_MAJOR

Sundials.SUNDIALS_VERSION_MINOR

Sundials.SUNDIALS_VERSION_PATCH

Sundials.SUNDenseLinearSolver

Sundials.SUNDenseMatrix

Sundials.SUNDenseMatrix_Cols

Sundials.SUNDenseMatrix_Column

Sundials.SUNDenseMatrix_Columns

Sundials.SUNDenseMatrix_Data

Sundials.SUNDenseMatrix_LData

Sundials.SUNDenseMatrix_Print

Sundials.SUNDenseMatrix_Rows

Sundials.SUNFALSE

Sundials.SUNKLU

Sundials.SUNKLUReInit

Sundials.SUNKLUSetOrdering

Sundials.SUNLINEARSOLVER_CUSTOM

Sundials.SUNLINEARSOLVER_DIRECT

Sundials.SUNLINEARSOLVER_ITERATIVE

Sundials.SUNLS_ATIMES_FAIL_REC

Sundials.SUNLS_ATIMES_FAIL_UNREC

Sundials.SUNLS_CONV_FAIL

Sundials.SUNLS_GS_FAIL

Sundials.SUNLS_ILL_INPUT

Sundials.SUNLS_LUFACT_FAIL

Sundials.SUNLS_MEM_FAIL

Sundials.SUNLS_MEM_NULL

Sundials.SUNLS_PACKAGE_FAIL_REC

Sundials.SUNLS_PACKAGE_FAIL_UNREC

Sundials.SUNLS_PSET_FAIL_REC

Sundials.SUNLS_PSET_FAIL_UNREC

Sundials.SUNLS_PSOLVE_FAIL_REC

Sundials.SUNLS_PSOLVE_FAIL_UNREC

Sundials.SUNLS_QRFACT_FAIL

Sundials.SUNLS_QRSOL_FAIL

Sundials.SUNLS_RES_REDUCED

Sundials.SUNLS_SUCCESS

Sundials.SUNLinSolFree

Sundials.SUNLinSolFree_Band

Sundials.SUNLinSolFree_Dense

Sundials.SUNLinSolFree_KLU

Sundials.SUNLinSolFree_PCG

Sundials.SUNLinSolFree_SPBCGS

Sundials.SUNLinSolFree_SPFGMR

Sundials.SUNLinSolFree_SPGMR

Sundials.SUNLinSolFree_SPTFQMR

Sundials.SUNLinSolGetType

Sundials.SUNLinSolGetType_Band

Sundials.SUNLinSolGetType_Dense

Sundials.SUNLinSolGetType_KLU

Sundials.SUNLinSolGetType_PCG

Sundials.SUNLinSolGetType_SPBCGS

Sundials.SUNLinSolGetType_SPFGMR

Sundials.SUNLinSolGetType_SPGMR

Sundials.SUNLinSolGetType_SPTFQMR

Sundials.SUNLinSolInitialize

Sundials.SUNLinSolInitialize_Band

Sundials.SUNLinSolInitialize_Dense

Sundials.SUNLinSolInitialize_KLU

Sundials.SUNLinSolInitialize_PCG

Sundials.SUNLinSolInitialize_SPBCGS

Sundials.SUNLinSolInitialize_SPFGMR

Sundials.SUNLinSolInitialize_SPGMR

Sundials.SUNLinSolInitialize_SPTFQMR

Sundials.SUNLinSolLastFlag

Sundials.SUNLinSolLastFlag_Band

Sundials.SUNLinSolLastFlag_Dense

Sundials.SUNLinSolLastFlag_KLU

Sundials.SUNLinSolLastFlag_PCG

Sundials.SUNLinSolLastFlag_SPBCGS

Sundials.SUNLinSolLastFlag_SPFGMR

Sundials.SUNLinSolLastFlag_SPGMR

Sundials.SUNLinSolLastFlag_SPTFQMR

Sundials.SUNLinSolNumIters

Sundials.SUNLinSolNumIters_PCG

Sundials.SUNLinSolNumIters_SPBCGS

Sundials.SUNLinSolNumIters_SPFGMR

Sundials.SUNLinSolNumIters_SPGMR

Sundials.SUNLinSolNumIters_SPTFQMR

Sundials.SUNLinSolResNorm

Sundials.SUNLinSolResNorm_PCG

Sundials.SUNLinSolResNorm_SPBCGS

Sundials.SUNLinSolResNorm_SPFGMR

Sundials.SUNLinSolResNorm_SPGMR

Sundials.SUNLinSolResNorm_SPTFQMR

Sundials.SUNLinSolResid

Sundials.SUNLinSolResid_PCG

Sundials.SUNLinSolResid_SPBCGS

Sundials.SUNLinSolResid_SPFGMR

Sundials.SUNLinSolResid_SPGMR

Sundials.SUNLinSolResid_SPTFQMR

Sundials.SUNLinSolSetATimes

Sundials.SUNLinSolSetATimes_PCG

Sundials.SUNLinSolSetATimes_SPBCGS

Sundials.SUNLinSolSetATimes_SPFGMR

Sundials.SUNLinSolSetATimes_SPGMR

Sundials.SUNLinSolSetATimes_SPTFQMR

Sundials.SUNLinSolSetPreconditioner

Sundials.SUNLinSolSetPreconditioner_PCG

Sundials.SUNLinSolSetPreconditioner_SPBCGS

Sundials.SUNLinSolSetPreconditioner_SPFGMR

Sundials.SUNLinSolSetPreconditioner_SPGMR

Sundials.SUNLinSolSetPreconditioner_SPTFQMR

Sundials.SUNLinSolSetScalingVectors

Sundials.SUNLinSolSetScalingVectors_PCG

Sundials.SUNLinSolSetScalingVectors_SPBCGS

Sundials.SUNLinSolSetScalingVectors_SPFGMR

Sundials.SUNLinSolSetScalingVectors_SPGMR

Sundials.SUNLinSolSetScalingVectors_SPTFQMR

Sundials.SUNLinSolSetup

Sundials.SUNLinSolSetup_Band

Sundials.SUNLinSolSetup_Dense

Sundials.SUNLinSolSetup_KLU

Sundials.SUNLinSolSetup_PCG

Sundials.SUNLinSolSetup_SPBCGS

Sundials.SUNLinSolSetup_SPFGMR

Sundials.SUNLinSolSetup_SPGMR

Sundials.SUNLinSolSetup_SPTFQMR

Sundials.SUNLinSolSolve

Sundials.SUNLinSolSolve_Band

Sundials.SUNLinSolSolve_Dense

Sundials.SUNLinSolSolve_KLU

Sundials.SUNLinSolSolve_PCG

Sundials.SUNLinSolSolve_SPBCGS

Sundials.SUNLinSolSolve_SPFGMR

Sundials.SUNLinSolSolve_SPGMR

Sundials.SUNLinSolSolve_SPTFQMR

Sundials.SUNLinSolSpace

Sundials.SUNLinSolSpace_Band

Sundials.SUNLinSolSpace_Dense

Sundials.SUNLinSolSpace_KLU

Sundials.SUNLinSolSpace_PCG

Sundials.SUNLinSolSpace_SPBCGS

Sundials.SUNLinSolSpace_SPFGMR

Sundials.SUNLinSolSpace_SPGMR

Sundials.SUNLinSolSpace_SPTFQMR

Sundials.SUNLinearSolver

Sundials.SUNLinearSolverContent_Band

Sundials.SUNLinearSolverContent_Dense

Sundials.SUNLinearSolverContent_PCG

Sundials.SUNLinearSolverContent_SPBCGS

Sundials.SUNLinearSolverContent_SPFGMR

Sundials.SUNLinearSolverContent_SPGMR

Sundials.SUNLinearSolverContent_SPTFQMR

Sundials.SUNLinearSolver_Ops

Sundials.SUNLinearSolver_Type

Sundials.SUNMATRIX_BAND

Sundials.SUNMATRIX_CUSTOM

Sundials.SUNMATRIX_DENSE

Sundials.SUNMATRIX_SPARSE

Sundials.SUNMatClone

Sundials.SUNMatClone_Band

Sundials.SUNMatClone_Dense

Sundials.SUNMatClone_Sparse

Sundials.SUNMatCopy

Sundials.SUNMatCopy_Band

Sundials.SUNMatCopy_Dense

Sundials.SUNMatCopy_Sparse

Sundials.SUNMatDestroy

Sundials.SUNMatDestroy_Band

Sundials.SUNMatDestroy_Dense

Sundials.SUNMatDestroy_Sparse

Sundials.SUNMatGetID

Sundials.SUNMatGetID_Band

Sundials.SUNMatGetID_Dense

Sundials.SUNMatGetID_Sparse

Sundials.SUNMatMatvec

Sundials.SUNMatMatvec_Band

Sundials.SUNMatMatvec_Dense

Sundials.SUNMatMatvec_Sparse

Sundials.SUNMatScaleAdd

Sundials.SUNMatScaleAddI

Sundials.SUNMatScaleAddI_Band

Sundials.SUNMatScaleAddI_Dense

Sundials.SUNMatScaleAddI_Sparse

Sundials.SUNMatScaleAdd_Band

Sundials.SUNMatScaleAdd_Dense

Sundials.SUNMatScaleAdd_Sparse

Sundials.SUNMatSpace

Sundials.SUNMatSpace_Band

Sundials.SUNMatSpace_Dense

Sundials.SUNMatSpace_Sparse

Sundials.SUNMatZero

Sundials.SUNMatZero_Band

Sundials.SUNMatZero_Dense

Sundials.SUNMatZero_Sparse

Sundials.SUNMatrix

Sundials.SUNMatrixContent_Band

Sundials.SUNMatrixContent_Dense

Sundials.SUNMatrixContent_Sparse

Sundials.SUNMatrix_ID

Sundials.SUNMatrix_Ops

Sundials.SUNPCG

Sundials.SUNPCGSetMaxl

Sundials.SUNPCGSetPrecType

Sundials.SUNPCG_MAXL_DEFAULT

Sundials.SUNRabs

Sundials.SUNRexp

Sundials.SUNRpowerI

Sundials.SUNRpowerR

Sundials.SUNRsqrt

Sundials.SUNSPBCGS

Sundials.SUNSPBCGSSetMaxl

Sundials.SUNSPBCGSSetPrecType

Sundials.SUNSPBCGS_MAXL_DEFAULT

Sundials.SUNSPFGMR

Sundials.SUNSPFGMRSetGSType

Sundials.SUNSPFGMRSetMaxRestarts

Sundials.SUNSPFGMRSetPrecType

Sundials.SUNSPFGMR_GSTYPE_DEFAULT

Sundials.SUNSPFGMR_MAXL_DEFAULT

Sundials.SUNSPFGMR_MAXRS_DEFAULT

Sundials.SUNSPGMR

Sundials.SUNSPGMRSetGSType

Sundials.SUNSPGMRSetMaxRestarts

Sundials.SUNSPGMRSetPrecType

Sundials.SUNSPGMR_GSTYPE_DEFAULT

Sundials.SUNSPGMR_MAXL_DEFAULT

Sundials.SUNSPGMR_MAXRS_DEFAULT

Sundials.SUNSPTFQMR

Sundials.SUNSPTFQMRSetMaxl

Sundials.SUNSPTFQMRSetPrecType

Sundials.SUNSPTFQMR_MAXL_DEFAULT

Sundials.SUNSparseFromBandMatrix

Sundials.SUNSparseFromDenseMatrix

Sundials.SUNSparseMatrix

Sundials.SUNSparseMatrix_Columns

Sundials.SUNSparseMatrix_Data

Sundials.SUNSparseMatrix_IndexPointers

Sundials.SUNSparseMatrix_IndexValues

Sundials.SUNSparseMatrix_NNZ

Sundials.SUNSparseMatrix_NP

Sundials.SUNSparseMatrix_Print

Sundials.SUNSparseMatrix_Realloc

Sundials.SUNSparseMatrix_Rows

Sundials.SUNSparseMatrix_SparseType

Sundials.SUNTRUE

Sundials.SecondOrderODEProblem

Sundials.SlsMat

Sundials.SparseMatrix

Sundials.SpbcgMem

Sundials.SpbcgMemRec

Sundials.SpfgmrMem

Sundials.SpfgmrMemRec

Sundials.SpgmrMem

Sundials.SpgmrMemRec

Sundials.SplitFunction

Sundials.SplitODEProblem

Sundials.SplitSDEFunction

Sundials.SplitSDEProblem

Sundials.SptfqmrMem

Sundials.SptfqmrMemRec

Sundials.SteadyStateProblem

Sundials.SteadyStateSolution

Sundials.StiffnessChoice

Sundials.Sundials

Sundials.SundialsDAEAlgorithm

Sundials.SundialsHandle

Sundials.SundialsLinSol

Sundials.SundialsMatrix

Sundials.SundialsODEAlgorithm

Sundials.TRBDF2_3_3_2

Sundials.TimeChoiceIterator

Sundials.TwoPointBVProblem

Sundials.UNIT_ROUNDOFF

Sundials.UserFunctionAndData

Sundials.VERNER_8_5_6

Sundials.ZONNEVELD_5_3_4

Sundials._DlsMat

Sundials._N_VectorContent_Serial

Sundials._SUNLinearSolverContent_Band

Sundials._SUNLinearSolverContent_Dense

Sundials._SUNLinearSolverContent_PCG

Sundials._SUNLinearSolverContent_SPBCGS

Sundials._SUNLinearSolverContent_SPFGMR

Sundials._SUNLinearSolverContent_SPGMR

Sundials._SUNLinearSolverContent_SPTFQMR

Sundials._SUNMatrixContent_Band

Sundials._SUNMatrixContent_Dense

Sundials._SUNMatrixContent_Sparse

Sundials._SlsMat

Sundials._SpfgmrMemRec

Sundials._SpgmrMemRec

Sundials.__ARKDlsGetReturnFlagName

Sundials.__ARKDlsSetMassLinearSolver

Sundials.__ARKSpilsGetReturnFlagName

Sundials.__ARKSpilsSetMassLinearSolver

Sundials.__ARKode

Sundials.__ARKodeGetDky

Sundials.__ARKodeGetErrWeights

Sundials.__ARKodeGetEstLocalErrors

Sundials.__ARKodeGetResWeights

Sundials.__ARKodeGetReturnFlagName

Sundials.__ARKodeInit

Sundials.__ARKodeLoadButcherTable

Sundials.__ARKodeReInit

Sundials.__ARKodeResVtolerance

Sundials.__ARKodeResize

Sundials.__ARKodeRootInit

Sundials.__ARKodeSVtolerances

Sundials.__ARKodeSetARKTableNum

Sundials.__ARKodeSetARKTables

Sundials.__ARKodeSetAdaptivityMethod

Sundials.__ARKodeSetDenseOrder

Sundials.__ARKodeSetERKTable

Sundials.__ARKodeSetERKTableNum

Sundials.__ARKodeSetFixedPoint

Sundials.__ARKodeSetIRKTable

Sundials.__ARKodeSetIRKTableNum

Sundials.__ARKodeSetLinear

Sundials.__ARKodeSetMaxConvFails

Sundials.__ARKodeSetMaxErrTestFails

Sundials.__ARKodeSetMaxHnilWarns

Sundials.__ARKodeSetMaxNonlinIters

Sundials.__ARKodeSetMaxNumSteps

Sundials.__ARKodeSetMaxStepsBetweenLSet

Sundials.__ARKodeSetOrder

Sundials.__ARKodeSetPredictorMethod

Sundials.__ARKodeSetSmallNumEFails

Sundials.__CVBBDPrecInitB

Sundials.__CVBBDPrecReInitB

Sundials.__CVBandPrecInitB

Sundials.__CVDiagB

Sundials.__CVDiagGetReturnFlagName

Sundials.__CVDlsGetReturnFlagName

Sundials.__CVDlsSetJacFnB

Sundials.__CVDlsSetJacFnBS

Sundials.__CVDlsSetLinearSolverB

Sundials.__CVSpilsGetReturnFlagName

Sundials.__CVSpilsSetEpsLinB

Sundials.__CVSpilsSetJacTimesB

Sundials.__CVSpilsSetJacTimesBS

Sundials.__CVSpilsSetLinearSolverB

Sundials.__CVSpilsSetPreconditionerB

Sundials.__CVSpilsSetPreconditionerBS

Sundials.__CVode

Sundials.__CVodeAdjInit

Sundials.__CVodeB

Sundials.__CVodeCreate

Sundials.__CVodeCreateB

Sundials.__CVodeF

Sundials.__CVodeGetAdjCVodeBmem

Sundials.__CVodeGetAdjDataPointHermite

Sundials.__CVodeGetAdjDataPointPolynomial

Sundials.__CVodeGetAdjY

Sundials.__CVodeGetB

Sundials.__CVodeGetDky

Sundials.__CVodeGetErrWeights

Sundials.__CVodeGetEstLocalErrors

Sundials.__CVodeGetQuad

Sundials.__CVodeGetQuadB

Sundials.__CVodeGetQuadDky

Sundials.__CVodeGetQuadErrWeights

Sundials.__CVodeGetQuadSens1

Sundials.__CVodeGetQuadSensDky

Sundials.__CVodeGetQuadSensDky1

Sundials.__CVodeGetReturnFlagName

Sundials.__CVodeGetSens1

Sundials.__CVodeGetSensDky

Sundials.__CVodeGetSensDky1

Sundials.__CVodeInit

Sundials.__CVodeInitB

Sundials.__CVodeInitBS

Sundials.__CVodeQuadInit

Sundials.__CVodeQuadInitB

Sundials.__CVodeQuadInitBS

Sundials.__CVodeQuadReInit

Sundials.__CVodeQuadReInitB

Sundials.__CVodeQuadSStolerancesB

Sundials.__CVodeQuadSVtolerances

Sundials.__CVodeQuadSVtolerancesB

Sundials.__CVodeReInit

Sundials.__CVodeReInitB

Sundials.__CVodeRootInit

Sundials.__CVodeSStolerancesB

Sundials.__CVodeSVtolerances

Sundials.__CVodeSVtolerancesB

Sundials.__CVodeSensInit

Sundials.__CVodeSensInit1

Sundials.__CVodeSensReInit

Sundials.__CVodeSetInitStepB

Sundials.__CVodeSetIterType

Sundials.__CVodeSetIterTypeB

Sundials.__CVodeSetMaxConvFails

Sundials.__CVodeSetMaxErrTestFails

Sundials.__CVodeSetMaxHnilWarns

Sundials.__CVodeSetMaxNonlinIters

Sundials.__CVodeSetMaxNumSteps

Sundials.__CVodeSetMaxNumStepsB

Sundials.__CVodeSetMaxOrd

Sundials.__CVodeSetMaxOrdB

Sundials.__CVodeSetMaxStepB

Sundials.__CVodeSetMinStepB

Sundials.__CVodeSetQuadErrCon

Sundials.__CVodeSetQuadErrConB

Sundials.__CVodeSetQuadSensErrCon

Sundials.__CVodeSetSensDQMethod

Sundials.__CVodeSetSensErrCon

Sundials.__CVodeSetSensMaxNonlinIters

Sundials.__CVodeSetStabLimDet

Sundials.__CVodeSetStabLimDetB

Sundials.__CVodeSetUserDataB

Sundials.__IDAAdjInit

Sundials.__IDABBDPrecInitB

Sundials.__IDABBDPrecReInitB

Sundials.__IDACalcIC

Sundials.__IDACalcICB

Sundials.__IDACalcICBS

Sundials.__IDADlsGetReturnFlagName

Sundials.__IDADlsSetJacFnB

Sundials.__IDADlsSetJacFnBS

Sundials.__IDADlsSetLinearSolverB

Sundials.__IDAGetAdjDataPointHermite

Sundials.__IDAGetAdjDataPointPolynomial

Sundials.__IDAGetAdjIDABmem

Sundials.__IDAGetAdjY

Sundials.__IDAGetB

Sundials.__IDAGetConsistentIC

Sundials.__IDAGetConsistentICB

Sundials.__IDAGetDky

Sundials.__IDAGetErrWeights

Sundials.__IDAGetEstLocalErrors

Sundials.__IDAGetQuad

Sundials.__IDAGetQuadB

Sundials.__IDAGetQuadDky

Sundials.__IDAGetQuadErrWeights

Sundials.__IDAGetQuadSens1

Sundials.__IDAGetQuadSensDky

Sundials.__IDAGetQuadSensDky1

Sundials.__IDAGetReturnFlagName

Sundials.__IDAGetSens1

Sundials.__IDAGetSensDky

Sundials.__IDAGetSensDky1

Sundials.__IDAInit

Sundials.__IDAInitB

Sundials.__IDAInitBS

Sundials.__IDAQuadInit

Sundials.__IDAQuadInitB

Sundials.__IDAQuadInitBS

Sundials.__IDAQuadReInit

Sundials.__IDAQuadReInitB

Sundials.__IDAQuadSStolerancesB

Sundials.__IDAQuadSVtolerances

Sundials.__IDAQuadSVtolerancesB

Sundials.__IDAReInit

Sundials.__IDAReInitB

Sundials.__IDARootInit

Sundials.__IDASStolerancesB

Sundials.__IDASVtolerances

Sundials.__IDASVtolerancesB

Sundials.__IDASensInit

Sundials.__IDASensReInit

Sundials.__IDASetConstraints

Sundials.__IDASetConstraintsB

Sundials.__IDASetId

Sundials.__IDASetIdB

Sundials.__IDASetInitStepB

Sundials.__IDASetLineSearchOffIC

Sundials.__IDASetMaxBacksIC

Sundials.__IDASetMaxConvFails

Sundials.__IDASetMaxErrTestFails

Sundials.__IDASetMaxNonlinIters

Sundials.__IDASetMaxNumItersIC

Sundials.__IDASetMaxNumJacsIC

Sundials.__IDASetMaxNumSteps

Sundials.__IDASetMaxNumStepsB

Sundials.__IDASetMaxNumStepsIC

Sundials.__IDASetMaxOrd

Sundials.__IDASetMaxOrdB

Sundials.__IDASetMaxStepB

Sundials.__IDASetQuadErrCon

Sundials.__IDASetQuadErrConB

Sundials.__IDASetQuadSensErrCon

Sundials.__IDASetSensDQMethod

Sundials.__IDASetSensErrCon

Sundials.__IDASetSensMaxNonlinIters

Sundials.__IDASetSuppressAlg

Sundials.__IDASetSuppressAlgB

Sundials.__IDASetUserDataB

Sundials.__IDASolve

Sundials.__IDASolveB

Sundials.__IDASolveF

Sundials.__IDASpilsGetReturnFlagName

Sundials.__IDASpilsSetEpsLinB

Sundials.__IDASpilsSetIncrementFactorB

Sundials.__IDASpilsSetJacTimesB

Sundials.__IDASpilsSetJacTimesBS

Sundials.__IDASpilsSetLinearSolverB

Sundials.__IDASpilsSetPreconditionerB

Sundials.__IDASpilsSetPreconditionerBS

Sundials.__KINDlsGetReturnFlagName

Sundials.__KINGetReturnFlagName

Sundials.__KINInit

Sundials.__KINSetConstraints

Sundials.__KINSetEtaForm

Sundials.__KINSetMAA

Sundials.__KINSetMaxBetaFails

Sundials.__KINSetMaxSetupCalls

Sundials.__KINSetMaxSubSetupCalls

Sundials.__KINSetNoInitSetup

Sundials.__KINSetNoMinEps

Sundials.__KINSetNoResMon

Sundials.__KINSetNumMaxIters

Sundials.__KINSetPrintLevel

Sundials.__KINSetSysFunc

Sundials.__KINSol

Sundials.__KINSpilsGetReturnFlagName

Sundials.__N_VAbs

Sundials.__N_VAbs_Serial

Sundials.__N_VAddConst

Sundials.__N_VAddConst_Serial

Sundials.__N_VClone

Sundials.__N_VCloneEmpty

Sundials.__N_VCloneEmptyVectorArray

Sundials.__N_VCloneEmpty_Serial

Sundials.__N_VCloneVectorArray

Sundials.__N_VCloneVectorArrayEmpty_Serial

Sundials.__N_VCloneVectorArray_Serial

Sundials.__N_VClone_Serial

Sundials.__N_VCompare

Sundials.__N_VCompare_Serial

Sundials.__N_VConst

Sundials.__N_VConst_Serial

Sundials.__N_VConstrMask

Sundials.__N_VConstrMask_Serial

Sundials.__N_VDestroy

Sundials.__N_VDestroyVectorArray

Sundials.__N_VDestroyVectorArray_Serial

Sundials.__N_VDestroy_Serial

Sundials.__N_VDiv

Sundials.__N_VDiv_Serial

Sundials.__N_VDotProd

Sundials.__N_VDotProd_Serial

Sundials.__N_VGetArrayPointer

Sundials.__N_VGetArrayPointer_Serial

Sundials.__N_VGetLength_Serial

Sundials.__N_VGetVectorID

Sundials.__N_VGetVectorID_Serial

Sundials.__N_VInv

Sundials.__N_VInvTest

Sundials.__N_VInvTest_Serial

Sundials.__N_VInv_Serial

Sundials.__N_VL1Norm

Sundials.__N_VL1Norm_Serial

Sundials.__N_VLinearSum

Sundials.__N_VLinearSum_Serial

Sundials.__N_VMaxNorm

Sundials.__N_VMaxNorm_Serial

Sundials.__N_VMin

Sundials.__N_VMinQuotient

Sundials.__N_VMinQuotient_Serial

Sundials.__N_VMin_Serial

Sundials.__N_VPrintFile_Serial

Sundials.__N_VPrint_Serial

Sundials.__N_VProd

Sundials.__N_VProd_Serial

Sundials.__N_VScale

Sundials.__N_VScale_Serial

Sundials.__N_VSetArrayPointer

Sundials.__N_VSetArrayPointer_Serial

Sundials.__N_VSpace

Sundials.__N_VSpace_Serial

Sundials.__N_VWL2Norm

Sundials.__N_VWL2Norm_Serial

Sundials.__N_VWrmsNorm

Sundials.__N_VWrmsNormMask

Sundials.__N_VWrmsNormMask_Serial

Sundials.__N_VWrmsNorm_Serial

Sundials.__SUNBandLinearSolver

Sundials.__SUNDIALSGetVersion

Sundials.__SUNDIALSGetVersionNumber

Sundials.__SUNDenseLinearSolver

Sundials.__SUNKLU

Sundials.__SUNKLUReInit

Sundials.__SUNKLUSetOrdering

Sundials.__SUNLinSolSetScalingVectors

Sundials.__SUNLinSolSetScalingVectors_PCG

Sundials.__SUNLinSolSetScalingVectors_SPBCGS

Sundials.__SUNLinSolSetScalingVectors_SPFGMR

Sundials.__SUNLinSolSetScalingVectors_SPGMR

Sundials.__SUNLinSolSetScalingVectors_SPTFQMR

Sundials.__SUNLinSolSolve

Sundials.__SUNLinSolSolve_Band

Sundials.__SUNLinSolSolve_Dense

Sundials.__SUNLinSolSolve_KLU

Sundials.__SUNLinSolSolve_PCG

Sundials.__SUNLinSolSolve_SPBCGS

Sundials.__SUNLinSolSolve_SPFGMR

Sundials.__SUNLinSolSolve_SPGMR

Sundials.__SUNLinSolSolve_SPTFQMR

Sundials.__SUNMatMatvec

Sundials.__SUNMatMatvec_Band

Sundials.__SUNMatMatvec_Dense

Sundials.__SUNMatMatvec_Sparse

Sundials.__SUNPCG

Sundials.__SUNPCGSetMaxl

Sundials.__SUNPCGSetPrecType

Sundials.__SUNRpowerI

Sundials.__SUNSPBCGS

Sundials.__SUNSPBCGSSetMaxl

Sundials.__SUNSPBCGSSetPrecType

Sundials.__SUNSPFGMR

Sundials.__SUNSPFGMRSetGSType

Sundials.__SUNSPFGMRSetMaxRestarts

Sundials.__SUNSPFGMRSetPrecType

Sundials.__SUNSPGMR

Sundials.__SUNSPGMRSetGSType

Sundials.__SUNSPGMRSetMaxRestarts

Sundials.__SUNSPGMRSetPrecType

Sundials.__SUNSPTFQMR

Sundials.__SUNSPTFQMRSetMaxl

Sundials.__SUNSPTFQMRSetPrecType

Sundials.__SUNSparseFromBandMatrix

Sundials.__SUNSparseFromDenseMatrix

Sundials.__SUNSparseMatrix

Sundials.__builtin_va_list

Sundials.__init__

Sundials._generic_N_Vector

Sundials._generic_N_Vector_Ops

Sundials._generic_SUNLinearSolver

Sundials._generic_SUNLinearSolver_Ops

Sundials._generic_SUNMatrix

Sundials._generic_SUNMatrix_Ops

Sundials.add_saveat!

Sundials.add_tstop!

Sundials.addat!

Sundials.addat_non_user_cache!

Sundials.addsteps!

Sundials.asarray

Sundials.auto_dt_reset!

Sundials.change_t_via_interpolation!

Sundials.check_deps

Sundials.check_error

Sundials.check_keywords

Sundials.cvode — Function.

cvode(f, y0::Vector{Float64}, t::Vector{Float64}, userdata::Any=nothing; integrator=:BDF, reltol::Float64=1e-3, abstol::Float64=1e-6, callback=(mem,t,y)->true)

f, Function of the form f(t, y::Vector{Float64}, yp::Vector{Float64}) where y is the input state vector, and yp is the output vector of time derivatives for the states y

y0, Vector of initial values

t, Vector of time values at which to record integration results

integrator, the chosen integration algorithm. Default is :BDF , other option is :Adams

reltol, Relative Tolerance to be used (default=1e-3)

abstol, Absolute Tolerance to be used (default=1e-6)

callback, Callback function of the form callback(mem, t_k::Float64, y_k::Vector{Float64})::Bool where mem is the integrator and t_k/y_k the time/state at timestep k. A return value of false exits the integrator at timestep k and returns the all the timesteps preceding k. (default=(mem,tk,yk)->true)

return: a solution matrix with time steps in t along rows and state variable y along columns

Sundials.cvode!

Sundials.cvodefun

Sundials.cvodefunjac

Sundials.cvodefunjac2

Sundials.cvodejac

Sundials.deleteat!

Sundials.deleteat_non_user_cache!

Sundials.depsfile

Sundials.du_cache

Sundials.eval

Sundials.full_cache

Sundials.get_dt

Sundials.get_du

Sundials.get_du!

Sundials.get_proposed_dt

Sundials.get_tmp_cache

Sundials.handle_callback_modifiers!

Sundials.handle_callbacks!

Sundials.handle_tstop!

Sundials.has_exp

Sundials.has_expmv

Sundials.has_expmv!

Sundials.has_ldiv

Sundials.has_ldiv!

Sundials.has_mul

Sundials.has_mul!

Sundials.idajac

Sundials.idasol — Function.

idasol(f, y0::Vector{Float64}, yp0::Vector{Float64}, t::Vector{Float64}, userdata::Any=nothing; reltol::Float64=1e-3, abstol::Float64=1e-6, diffstates::Union{Vector{Bool},Void}=nothing)

f, Function of the form f(t, y::Vector{Float64}, yp::Vector{Float64}, r::Vector{Float64})whereyandypare the input state and derivative vectors, andr` is the output residual vector

y0, Vector of initial values

yp0, Vector of initial values of the derivatives

reltol, Relative Tolerance to be used (default=1e-3)

abstol, Absolute Tolerance to be used (default=1e-6)

diffstates, Boolean vector, true for the positions such that r depends on yp[k]

return: (y,yp) two solution matrices representing the states and state derivatives with time steps in t along rows and state variable y or yp along columns

Sundials.idasolfun

Sundials.include

Sundials.init

Sundials.initialize!

Sundials.interpret_sundials_retcode

Sundials.intervals

Sundials.is_constant

Sundials.isdiscrete

Sundials.isinplace

Sundials.kinsol

Sundials.kinsolfun

Sundials.libamd

Sundials.libbtf

Sundials.libcolamd

Sundials.libklu

Sundials.libsuitesparseconfig

Sundials.libsundials_arkode

Sundials.libsundials_cvode

Sundials.libsundials_cvodes

Sundials.libsundials_ida

Sundials.libsundials_idas

Sundials.libsundials_kinsol

Sundials.libsundials_nvecserial

Sundials.libsundials_sunlinsolband

Sundials.libsundials_sunlinsoldense

Sundials.libsundials_sunlinsolklu

Sundials.libsundials_sunlinsolpcg

Sundials.libsundials_sunlinsolspbcgs

Sundials.libsundials_sunlinsolspfgmr

Sundials.libsundials_sunlinsolspgmr

Sundials.libsundials_sunlinsolsptfqmr

Sundials.libsundials_sunmatrixband

Sundials.libsundials_sunmatrixdense

Sundials.libsundials_sunmatrixsparse

Sundials.linear_solver

Sundials.method_choice

Sundials.null_error_handler

Sundials.nvector

Sundials.nvlength

Sundials.old_cfunction

Sundials.rand_cache

Sundials.ratenoise_cache

Sundials.realtype

Sundials.reeval_internals_due_to_modification!

Sundials.reinit!

Sundials.release_handle

Sundials.remake

Sundials.resize!

Sundials.resize_non_user_cache!

Sundials.save_value!

Sundials.savevalues!

Sundials.set_abstol!

Sundials.set_proposed_dt!

Sundials.set_reltol!

Sundials.set_stop_time

Sundials.set_t!

Sundials.set_u!

Sundials.solve

Sundials.solve!

Sundials.solver_step

Sundials.step!

Sundials.sunindextype

Sundials.terminate!

Sundials.tstop_saveat_disc_handling

Sundials.tuples

Sundials.u_cache

Sundials.u_modified!

Sundials.update_coefficients

Sundials.update_coefficients!

Sundials.user_cache

Sundials.warn_compat

Sundials.warnida

Sundials.warnkeywords

Sundials.warnlist