Docstrings

BayesNets.+

BayesNets.-

BayesNets.@csv2_str

BayesNets.@csv_str

BayesNets.@pack_GibbsSamplingFull

BayesNets.@pack_GibbsSamplingFull!

BayesNets.@pack_GibbsSamplingNodewise

BayesNets.@pack_GibbsSamplingNodewise!

BayesNets.@pack_LikelihoodWeightingInference

BayesNets.@pack_LikelihoodWeightingInference!

BayesNets.@pack_LoopyBelief

BayesNets.@pack_LoopyBelief!

BayesNets.@tsv_str

BayesNets.@unpack_GibbsSamplingFull

BayesNets.@unpack_GibbsSamplingNodewise

BayesNets.@unpack_LikelihoodWeightingInference

BayesNets.@unpack_LoopyBelief

BayesNets.@wsv_str

BayesNets.AbstractCategoricalArray

BayesNets.AbstractCategoricalMatrix

BayesNets.AbstractCategoricalVector

BayesNets.AbstractDataFrame

BayesNets.AbstractMissingCategoricalArray

BayesNets.AbstractMissingCategoricalMatrix

BayesNets.AbstractMissingCategoricalVector

BayesNets.AbstractMixtureModel

BayesNets.AbstractMvNormal

BayesNets.Arcsine

BayesNets.Assignment

BayesNets.BDeuPrior — Type.

Assigns equal scores to Markov equivalent structures

α_ijk = x/{q_i * r_i} for each j, k and some given x

see DMU section 2.4.3

BayesNets.BayesNet

BayesNets.BayesNetSampler — Type.

Abstract type for sampling with:

Random.rand(BayesNet, BayesNetSampler)

Random.rand(BayesNet, BayesNetSampler, nsamples)

Random.rand!(Assignment, BayesNet, BayesNetSampler)

BayesNets.BayesNets

BayesNets.Bernoulli

BayesNets.Beta

BayesNets.BetaBinomial

BayesNets.BetaPrime

BayesNets.Binomial

BayesNets.Biweight

BayesNets.CPD

BayesNets.CPDs

BayesNets.Categorical

BayesNets.CategoricalArray

BayesNets.CategoricalArrays

BayesNets.CPDs.CategoricalCPD — Type.

A categorical distribution, P(x|parents(x)) where all parents are discrete integers 1:N.

The ordering of distributions array follows the convention in Decision Making Under Uncertainty. Suppose a variable has three discrete parents. The first parental instantiation assigns all parents to their first bin. The second will assign the first parent (as defined in parents) to its second bin and the other parents to their first bin. The sequence continues until all parents are instantiated to their last bins.

This is equivalent to:

X,Y,Z 1,1,1 2,1,1 1,2,1 2,2,1 1,1,2 ...

BayesNets.CategoricalMatrix

BayesNets.CategoricalPool

BayesNets.CategoricalString

BayesNets.CategoricalValue

BayesNets.CategoricalVector

BayesNets.Cauchy

BayesNets.Chi

BayesNets.Chisq

BayesNets.CPDs.ConditionalLinearGaussianCPD — Type.

A conditional linear Gaussian CPD, always returns a Normal{Float64}

This is a combination of the CategoricalCPD and the LinearGaussianCPD.
For a variable with N discrete parents and M continuous parents, it will construct
a linear gaussian distribution for all M parents for each discrete instantiation.

                  { Normal(μ=a₁×continuous_parents(x) + b₁, σ₁) for discrete instantiation 1
P(x|parents(x)) = { Normal(μ=a₂×continuous_parents(x) + b₂, σ₂) for discrete instantiation 2
                  { ...

BayesNets.Continuous

BayesNets.ContinuousDistribution

BayesNets.ContinuousMatrixDistribution

BayesNets.ContinuousMultivariateDistribution

BayesNets.ContinuousUnivariateDistribution

BayesNets.Cosine

BayesNets.DAG

BayesNets.DataFrame

BayesNets.DataFrameRow

BayesNets.DataFrames

BayesNets.DiagNormal

BayesNets.DiagNormalCanon

BayesNets.DirectSampler — Type.

Straightforward sampling from a BayesNet. The default sampler.

BayesNets.Dirichlet

BayesNets.DirichletMultinomial

BayesNets.DirichletPrior — Type.

Baysian Structure learning seeks to maximize P(G|D) In the Bayesian fashion, we can provide a prior over the parameters in our learning network. This is described using a Dirichlet Prior.

BayesNets.Discrete

BayesNets.DiscreteBayesNet — Type.

DiscreteBayesNets are Bayesian Networks where every variable is an integer within 1:Nᵢ and every distribution is Categorical.

This representation is very common, and allows for the use of factors, for example in Probabilistic Graphical Models by Koller and Friedman

BayesNets.CPDs.DiscreteCPD — Type.

A categorical distribution, P(x|parents(x)) where all parents are discrete integers 1:N.

The ordering of distributions array follows the convention in Decision Making Under Uncertainty. Suppose a variable has three discrete parents. The first parental instantiation assigns all parents to their first bin. The second will assign the first parent (as defined in parents) to its second bin and the other parents to their first bin. The sequence continues until all parents are instantiated to their last bins.

This is equivalent to:

X,Y,Z 1,1,1 2,1,1 1,2,1 2,2,1 1,1,2 ...

BayesNets.DiscreteDistribution

BayesNets.DiscreteMatrixDistribution

BayesNets.DiscreteMultivariateDistribution

BayesNets.DiscreteUniform

BayesNets.DiscreteUnivariateDistribution

BayesNets.Distribution

BayesNets.Distributions

BayesNets.DoubleExponential

BayesNets.EdgeworthMean

BayesNets.EdgeworthSum

BayesNets.EdgeworthZ

BayesNets.EmpiricalUnivariateDistribution

BayesNets.Epanechnikov

BayesNets.Erlang

BayesNets.Estimator

BayesNets.ExactInference — Type.

Exact inference using factors and variable eliminations

BayesNets.Exponential

BayesNets.FDist

BayesNets.Factor — Type.

Factor(dims, potential)

Create a Factor corresponding to the potential.

BayesNets.FisherNoncentralHypergeometric

BayesNets.Frechet

BayesNets.FullNormal

BayesNets.FullNormalCanon

BayesNets.FunctionalCPD

BayesNets.Gamma

BayesNets.GeneralizedExtremeValue

BayesNets.GeneralizedPareto

BayesNets.Geometric

BayesNets.GibbsSampler — Type.

The GibbsSampler type houses the parameters of the Gibbs sampling algorithm. The parameters are defined below:

burnin: The first burnin samples will be discarded. They will not be returned. The thinning parameter does not affect the burn in period. This is used to ensure that the Gibbs sampler converges to the target stationary distribution before actual samples are drawn.

thinning: For every thinning + 1 number of samples drawn, only the last is kept. Thinning is used to reduce autocorrelation between samples. Thinning is not used during the burn in period. e.g. If thinning is 1, samples will be drawn in groups of two and only the second sample will be in the output.

timelimit: The number of milliseconds to run the algorithm. The algorithm will return the samples it has collected when either nsamples samples have been collected or timelimit milliseconds have passed. If time_limit is null then the algorithm will run until nsamples have been collected. This means it is possible that zero samples are returned.

erroriftimeout: If erroriftimeout is true and the timelimit expires, an error will be raised. If erroriftimeout is false and the time limit expires, the samples that have been collected so far will be returned. This means it is possible that zero samples are returned. Burn in samples will not be returned. If time_limit is null, this parameter does nothing.

consistent_with: the assignment that all samples must be consistent with (ie, Assignment(:A=>1) means all samples must have :A=1). Use to sample conditional distributions.

maxcachesize: If null, cache as much as possible, otherwise cache at most "maxcachesize" distributions

variableorder: variableorder determines the order of variables changed when generating a new sample. If null use a random order for every sample (this is different from updating the variables at random). Otherwise should be a list containing all the variables in the order they should be updated.

initial_sample: The inital assignment to variables to use. If null, the initial sample is chosen by briefly using a LikelihoodWeightedSampler.

BayesNets.GibbsSamplerState — Type.

Used to cache various things the Gibbs sampler needs

BayesNets.GibbsSamplingFull — Type.

infer(im, inf)

Run Gibbs sampling for N iterations. Each iteration changes all nodes. Discareds first burn_in samples and keeps only the thin-th sample. Ex, if thin=3, will discard the first two samples and keep the third.

source

BayesNets.GibbsSamplingNodewise — Type.

infer(GibbsSampling, state::Assignment, InferenceState)

Run Gibbs sampling for N iterations. Each iteration changes one node.

Discareds first burn_in samples and keeps only the thin-th sample. Ex, if thin=3, will discard the first two samples and keep the third.

source

BayesNets.CPDs.ProbabilisticGraphicalModels.GraphSearchStrategy — Type.

GraphSearchStrategy

An abstract type which defines a graph search strategy for learning probabilistic graphical model structures These allow: fit(::Type{ProbabilisticGraphicalModel}, data, GraphSearchStrategy)

BayesNets.GreedyHillClimbing

BayesNets.GreedyThickThinning

BayesNets.GroupApplied

BayesNets.GroupedDataFrame

BayesNets.Gumbel

BayesNets.Hypergeometric

BayesNets.CPDs.ProbabilisticGraphicalModels.InferenceMethod — Type.

Abstract type for probability inference

BayesNets.CPDs.ProbabilisticGraphicalModels.InferenceState — Type.

Type for capturing the inference state

BayesNets.InverseGamma

BayesNets.InverseGaussian

BayesNets.InverseWishart

BayesNets.IsoNormal

BayesNets.IsoNormalCanon

BayesNets.K2GraphSearch — Type.

K2GraphSearch

A GraphSearchStrategy following the K2 algorithm. Takes polynomial time to find the optimal structure assuming a topological variable ordering.

BayesNets.KSDist

BayesNets.KSOneSided

BayesNets.Kolmogorov

BayesNets.Laplace

BayesNets.LevelsException

BayesNets.Levy

BayesNets.LikelihoodWeightedSampler — Type.

Likelihood Weighted Sampling

BayesNets.LikelihoodWeightingInference — Type.

Approximates p(query|evidence) with N weighted samples using likelihood weighted sampling

source

BayesNets.CPDs.LinearGaussianCPD — Type.

A linear Gaussian CPD, always returns a Normal

Assumes that target and all parents can be converted to Float64 (ie, are numeric)

P(x|parents(x)) = Normal(μ=a×parents(x) + b, σ)

BayesNets.LocationScale

BayesNets.LogNormal

BayesNets.Logistic

BayesNets.LoopyBelief — Type.

Loopy belief propogation for a network.

Early stopping if change is messages < tol for `itersforconvergence' iterations. For no stopping, use tol < 0.

source

BayesNets.MLEstimator

BayesNets.MatrixDistribution

BayesNets.Matrixvariate

BayesNets.Missing

BayesNets.MissingCategoricalArray

BayesNets.MissingCategoricalMatrix

BayesNets.MissingCategoricalVector

BayesNets.MissingException

BayesNets.Missings

BayesNets.MixtureModel

BayesNets.Multinomial

BayesNets.Multivariate

BayesNets.MultivariateDistribution

BayesNets.MultivariateMixture

BayesNets.MultivariateNormal

BayesNets.MvLogNormal

BayesNets.MvNormal

BayesNets.MvNormalCanon

BayesNets.MvNormalKnownCov

BayesNets.MvTDist

BayesNets.NamedCategorical

BayesNets.NegativeBayesianInformationCriterion — Type.

NegativeBayesianInformationCriterion

A ScoringFunction for the negative Bayesian information criterion.

BIC = -2⋅L + k⋅ln(n)

   L - the log likelihood of the data under the cpd
   k - the number of free parameters to be estimated
   n - the sample size

BayesNets.NegativeBinomial

BayesNets.NodeName

BayesNets.NodeNameUnion

BayesNets.NodeNames

BayesNets.NonMatrixDistribution

BayesNets.NoncentralBeta

BayesNets.NoncentralChisq

BayesNets.NoncentralF

BayesNets.NoncentralHypergeometric

BayesNets.NoncentralT

BayesNets.Normal

BayesNets.NormalCanon

BayesNets.NormalInverseGaussian

BayesNets.OrderedLevelsException

BayesNets.ParentFunctionalCPD

BayesNets.Pareto

BayesNets.Poisson

BayesNets.PoissonBinomial

BayesNets.ProbabilisticGraphicalModel

BayesNets.CPDs.ProbabilisticGraphicalModels — Module.

Provides a basic interface for defining and working with probabilistic graphical models

BayesNets.QQPair

BayesNets.Rayleigh

BayesNets.RealInterval

BayesNets.RejectionSampler — Type.

Rejection Sampling in which the assignments are forced to be consistent with the provided values. Each sampler is attempted at most max_nsamples times before returning an empty assignment.

BayesNets.Sampleable

BayesNets.CPDs.ProbabilisticGraphicalModels.Sampler — Type.

Abstract type for sampling with Base.rand(ProbabilisticGraphicalModel, Sampler, nsamples) Base.rand!(Assignment, ProbabilisticGraphicalModel, Sampler) Base.rand(ProbabilisticGraphicalModel, Sampler)

BayesNets.ScanGreedyHillClimbing

BayesNets.ScoreComponentCache — Type.

ScoreComponentCache

Used to store scores in a priority queue such that graph search algorithms know when a particular construction has already been made. cacheⱼ for the ith variable with parents parents

BayesNets.ScoringFunction — Type.

ScoringFunction

An abstract type for which subtypes allow extracting CPD score components, which are to be maximized: score_component(::ScoringFunction, cpd::CPD, data::DataFrame)

BayesNets.Semicircle

BayesNets.Skellam

BayesNets.CPDs.StaticCPD — Type.

A CPD for which the distribution never changes. target: name of the CPD's variable parents: list of parent variables. d: a Distributions.jl distribution

While a StaticCPD can have parents, their assignments will not affect the distribution.

BayesNets.SubDataFrame

BayesNets.SufficientStats

BayesNets.SymTriangularDist

BayesNets.TDist

BayesNets.Table

BayesNets.TriangularDist

BayesNets.Triweight

BayesNets.Truncated

BayesNets.TruncatedNormal

BayesNets.Uniform

BayesNets.UniformPrior — Type.

A uniform Dirichlet prior such that all α are the same

Defaults to the popular K2 prior, α = 1, which is similar to Laplace Smoothing

https://en.wikipedia.org/wiki/Additive_smoothing

BayesNets.Univariate

BayesNets.UnivariateDistribution

BayesNets.UnivariateGMM

BayesNets.UnivariateMixture

BayesNets.ValueSupport

BayesNets.VariateForm

BayesNets.VonMises

BayesNets.VonMisesFisher

BayesNets.WalleniusNoncentralHypergeometric

BayesNets.Weibull

BayesNets.Wishart

BayesNets.ZeroMeanDiagNormal

BayesNets.ZeroMeanDiagNormalCanon

BayesNets.ZeroMeanFullNormal

BayesNets.ZeroMeanFullNormalCanon

BayesNets.ZeroMeanIsoNormal

BayesNets.ZeroMeanIsoNormalCanon

BayesNets._build_dag_from_cpds

BayesNets._check_dims_valid

BayesNets._ckeck_dims_unique

BayesNets._enforce_topological_order

BayesNets._evidence_lambda — Function.

Get the lambda-message to itself for an evidence node. If it isn't an evidence node, this will break

BayesNets._get_dag

BayesNets._get_parent_indeces — Function.

score_component(a::ScoringFunction, cpd::CPD, data::DataFrame, cache::ScoreComponentCache)

As score_component(ScoringFunction, cpd, data), but returns pre-computed values from the cache if they exist, and populates the cache if they don't

BayesNets._get_parental_ncategories

BayesNets._init_gibbs_sample — Function.

_init_gibbs_sample(bn, evidence)

A random sample of non-evidence nodes uniformly over their domain

BayesNets._reddim

BayesNets._translate_index

BayesNets.add_edge!

BayesNets.adding_edge_preserves_acyclicity

BayesNets.aggregate

BayesNets.allowmissing

BayesNets.allowmissing!

BayesNets.bayesian_score — Function.

bayesian_score(G::DAG, names::Vector{Symbol}, data::DataFrame[, ncategories::Vector{Int}[, prior::DirichletPrior]])

Compute the bayesian score for graph structure g, with the data in data. names containes a symbol corresponding to each vertex in g that is the name of a column in data. ncategories is a vector of the number of values that each variable in the Bayesian network can take.

Note that every entry in data must be an integer greater than 0

BayesNets.bayesian_score_component — Function.

Computes the Bayesian score component for the given target variable index and Dirichlet prior counts given in alpha

INPUT: i - index of the target variable parents - list of indeces of parent variables (should not contain self) r - list of instantiation counts accessed by variable index r[1] gives number of discrete states variable 1 can take on data - matrix of sufficient statistics / counts d[j,k] gives the number of times the target variable took on its kth instantiation given the jth parental instantiation

OUTPUT: the Bayesian score, Float64

BayesNets.bayesian_score_component_uniform

BayesNets.bayesian_score_components

BayesNets.binaryentropy

BayesNets.by

BayesNets.canonform

BayesNets.categorical

BayesNets.categorical!

BayesNets.ccdf

BayesNets.cdf

BayesNets.cf

BayesNets.cgf

BayesNets.children — Function.

Returns the children as a list of NodeNames

BayesNets.circvar

BayesNets.coalesce

BayesNets.colwise

BayesNets.combine

BayesNets.compact

BayesNets.complete_cases

BayesNets.complete_cases!

BayesNets.completecases

BayesNets.component

BayesNets.components

BayesNets.componentwise_logpdf

BayesNets.componentwise_pdf

BayesNets.compress

BayesNets.concentration

BayesNets.CPDs.ProbabilisticGraphicalModels.consistent — Function.

consistent(a::Assignment, b::Assignment)

True if all shared NodeNames have the same value

Base.convert — Function.

convert(DiscreteCPD, cpd)

Construct a Factor from a DiscreteCPD.

Convert a Factor to a DataFrame

BayesNets.cor

BayesNets.cov

BayesNets.cquantile

BayesNets.cumulant

BayesNets.cut

BayesNets.decompress

Base.delete! — Function.

delete!(bn::BayesNets, target::NodeName)

Removing cpds will alter the vertex indeces. In particular, removing the ith cpd will swap i and n and then remove n.

BayesNets.deletecols!

BayesNets.deleterows!

BayesNets.descendants

BayesNets.describe

BayesNets.dim

BayesNets.disallowmissing

BayesNets.disallowmissing!

BayesNets.CPDs.disttype — Function.

disttype(cpd::CPD)

Return the type of the CPD's distribution

BayesNets.dof

BayesNets.droplevels!

BayesNets.dropmissing

BayesNets.dropmissing!

BayesNets.duplicate — Function.

duplicate(A, dims)

Repeates an array only through higer dimensions dims.

Custom version of repeate, but only outer repetition, and only duplicates the array for the number of times specified in dims for dimensions greater than ndims(A). If dims is empty, returns a copy of A.

julia> duplicate(collect(1:3), 2)
3×2 Array{Int64,2}:
 1  1
 2  2
 3  3

julia> duplicate([1 3; 2 4], 3)
2×2×3 Array{Int64,3}:
[:, :, 1] =
 1  3
 2  4

[:, :, 2] =
 1  3
 2  4

[:, :, 3] =
 1  3
 2  4

BayesNets.eltypes

BayesNets.enforce_topological_order!

BayesNets.entropy

BayesNets.estimate

BayesNets.eval

BayesNets.eval_mb_cpd — Function.

eval_mb_cpd(node, ncategories, assignment, mb_cpds)

Return the potential of all instances of a node given its markove blanket as a WeightVec: P(node | panode) * Prod (c in children) P(c | pac)

Trys out all possible values of node (assumes categorical) Assignment should have values for all in the Markov blanket, including the variable itself.

BayesNets.expected_logdet

BayesNets.failprob

StatsBase.fit — Function.

fit(::Type{ProbabilisticGraphicalModel}, data::DataFrame, params::GraphSearchStrategy)

Runs the graph search algorithm to learn a probabilistic graphical model of the provided type from data.

fit(::Type{CPD}, data::DataFrame, target::NodeName, parents::NodeNames)

Construct a CPD for target by fitting it to the provided data

fit(::Type{BayesNet}, data, edges)

Fit a Bayesian Net whose variables are the columns in data and whose edges are given in edges

ex: fit(DiscreteBayesNet, data, (:A=>:B, :C=>B))

fit{C<:CPD}(::Type{BayesNet{C}}, ::DataFrame, ::GraphSearchStrategy)

Run the graph search algorithm defined by GraphSearchStrategy

takes a list of observations of assignments represented as a DataFrame or a set of data samples (without :p), takes the unique assignments, and estimates the associated probability of each assignment based on its frequency of occurrence.

BayesNets.fit_map

BayesNets.fit_map!

BayesNets.fit_mle

BayesNets.fit_mle!

BayesNets.freecumulant

BayesNets.get_asia_bn — Function.

An ergodic version of the asia network, with the E variable removed

Orignal network: Lauritzen, Steffen L. and David J. Spiegelhalter, 1988

BayesNets.get_finite_distribution! — Function.

Helper to sampleposteriorfinite

Modifies a and gss

BayesNets.get_mb_cpds — Function.

Get the cpd's of a node and its children

BayesNets.get_sat_fail_bn — Function.

Satellite failure network from DMU, pg 17

BayesNets.get_sprinkler_bn — Function.

The usual sprinkler problem

BayesNets.get_weighted_dataframe — Function.

A dataset of variable assignments is obtained with an additional column of weights in accordance with the likelihood of each assignment.

BayesNets.get_weighted_sample

BayesNets.get_weighted_sample! — Function.

Draw an assignment from the Bayesian network but set any variables in the evidence accordingly. Returns the assignment and the probability weighting associated with the evidence.

BayesNets.gibbs_sample — Function.

Implements Gibbs sampling. (https://en.wikipedia.org/wiki/Gibbs_sampling) For finite variables, the posterior distribution is sampled by building the exact distribution. For continuous variables, the posterior distribution is sampled using Metropolis Hastings MCMC. Discrete variables with infinite support are currently not supported. The Gibbs Sampler only supports CPDs that return Univariate Distributions. (CPD{D<:UnivariateDistribution})

bn:: A Bayesian Network to sample from. bn should only contain CPDs that return UnivariateDistributions.

nsamples: The number of samples to return.

burnin: The first burnin samples will be discarded. They will not be returned. The thinning parameter does not affect the burn in period. This is used to ensure that the Gibbs sampler converges to the target stationary distribution before actual samples are drawn.

thinning: For every thinning + 1 number of samples drawn, only the last is kept. Thinning is used to reduce autocorrelation between samples. Thinning is not used during the burn in period. e.g. If thinning is 1, samples will be drawn in groups of two and only the second sample will be in the output.

timelimit: The number of milliseconds to run the algorithm. The algorithm will return the samples it has collected when either nsamples samples have been collected or timelimit milliseconds have passed. If time_limit is null then the algorithm will run until nsamples have been collected. This means it is possible that zero samples are returned.

erroriftimeout: If erroriftimeout is true and the timelimit expires, an error will be raised. If erroriftimeout is false and the time limit expires, the samples that have been collected so far will be returned. This means it is possible that zero samples are returned. Burn in samples will not be returned. If time_limit is null, this parameter does nothing.

consistent_with: the assignment that all samples must be consistent with (ie, Assignment(:A=>1) means all samples must have :A=1). Use to sample conditional distributions.

maxcachesize: If null, cache as much as possible, otherwise cache at most "maxcachesize" distributions

variableorder: variableorder determines the order of variables changed when generating a new sample. If null use a random order for every sample (this is different from updating the variables at random). Otherwise should be a list containing all the variables in the order they should be updated.

initialsample: The inital assignment to variables to use. If null, the initial sample is chosen by briefly running rand(bn, getweighted_dataframe).

BayesNets.gibbs_sample_main_loop — Function.

The main loop associated with Gibbs sampling Returns a data frame with nsamples samples

Supports the various parameters supported by gibbssample Refer to gibbssample for parameter meanings

BayesNets.gradlogpdf

BayesNets.greedy_score

BayesNets.groupby

BayesNets.groupindices

BayesNets.groupvars

LightGraphs.has_edge — Function.

Whether the BayesNet contains the given edge

BayesNets.hasfinitesupport

BayesNets.head

BayesNets.include

BayesNets.index_data

BayesNets.CPDs.ProbabilisticGraphicalModels.infer — Function.

infer(InferenceMethod, InferenceState)

Infer p(query|evidence)

Approximates p(query|evidence) with nsamples likelihood weighted samples.

Since this uses a Factor, it is only efficient if the number of samples is (signifcantly) greater than the number of possible instantiations for the query variables

BayesNets.CPDs.infer_number_of_instantiations — Function.

infer_number_of_instantiations{I<:Int}(arr::AbstractVector{I})

Infer the number of instantiations, N, for a data type, assuming that it takes on the values 1:N

BayesNets.insert!

BayesNets.insertcols!

BayesNets.insupport

BayesNets.invalid_dims_error

BayesNets.invcov

BayesNets.invlogccdf

BayesNets.invlogcdf

BayesNets.invscale

BayesNets.CPDs.ProbabilisticGraphicalModels.is_independent — Function.

is_independent(PGM, x::NodeNames, y::NodeNames, given::NodeNames) Returns whether the set of node names x is d-separated from the set y given the set given

Returns whether the set of node names x is d-separated from the set y given the set given

BayesNets.isbounded

BayesNets.isleptokurtic

BayesNets.islowerbounded

BayesNets.ismesokurtic

BayesNets.ismissing

BayesNets.isordered

BayesNets.isplatykurtic

BayesNets.isprobvec

BayesNets.isupperbounded

BayesNets.kde

BayesNets.keys

BayesNets.kurtosis

Base.length — Function.

length(PGM)

Returns the number of variables in the probabilistic graphical model

Total number of elements in Factor (potential)

BayesNets.levels

BayesNets.levels!

BayesNets.location

BayesNets.location!

BayesNets.logccdf

BayesNets.logcdf

BayesNets.logdetcov

BayesNets.loglikelihood

Distributions.logpdf — Function.

The logpdf of a given assignment after conditioning on the values

The logpdf of a set of assignment after conditioning on the values

logpdf(cpd::CPD)

Condition and then return the logpdf

logpdf(cpd::CPD, data::DataFrame)

Return the logpdf across the dataset

The logpdf of a given assignment after conditioning on the values

BayesNets.logpdf!

BayesNets.map

BayesNets.mapcols

BayesNets.CPDs.ProbabilisticGraphicalModels.markov_blanket — Function.

markov_blanket(PGM) Returns the list of NodeNames forming the Markov blanket for the PGM

Return the children, parents, and parents of children (excluding target) as a Set of NodeNames

BayesNets.mean

BayesNets.meandir

BayesNets.meanform

BayesNets.meanlogx

BayesNets.median

BayesNets.melt

BayesNets.meltdf

BayesNets.merge!

BayesNets.meshgrid

BayesNets.mgf

BayesNets.missing

BayesNets.missings

BayesNets.mode

BayesNets.modes

BayesNets.moment

BayesNets.CPDs.name — Function.

name(cpd::CPD)

Return the NodeName for the variable this CPD is defined for.

BayesNets.names!

Distributions.ncategories — Function.

Distributions.ncategories(cpd::CategoricalCPD)

Return the number of categories for a cpd.

Distributions.ncategories(bn::DiscreteBayesNet, node::Symbol)

Return the number of categories for a node in the network.

BayesNets.ncol

BayesNets.ncomponents

BayesNets.ndgrid

BayesNets.ndgrid_fill! — Function.

???

LightGraphs.neighbors — Function.

Returns all neighbors as a list of NodeNames.

BayesNets.non_unique_dims_error

BayesNets.non_unique_states_error

BayesNets.nonunique

LinearAlgebra.normalize — Function.

normalize!(ϕ, dims; p=1)
normalize!(ϕ; p=1)

Return a normalized copy of the factor so all instances of dims have (or the entire factors has) p-norm of 1

BayesNets.not_in_factor_error

BayesNets.CPDs.nparams — Function.

nparams(cpd::CPD)

Return the number of free parameters that needed to be estimated for the CPD

BayesNets.nrow

BayesNets.nsamples

BayesNets.ntrials

BayesNets.nullable!

BayesNets.order

BayesNets.ordered

BayesNets.ordered!

BayesNets.params

BayesNets.params!

BayesNets.CPDs.parentless — Function.

parentless(cpd::CPD)

Return whether this CPD has parents.

BayesNets.CPDs.parents — Function.

parents(cpd::CPD)

Return the parents for this CPD as a vector of NodeName.

Returns the parents as a list of NodeNames

BayesNets.partype

BayesNets.passmissing

BayesNets.pattern — Function.

pattern(ϕ, [dims])

Return an array with the pattern of each dimension's state for all possible instances

Distributions.pdf — Function.

The pdf of a given assignment after conditioning on the values

The pdf of a set of assignments after conditioning on the values

pdf(cpd::CPD)

Condition and then return the pdf

pdf(cpd::CPD, data::DataFrame)

Return the pdf across the dataset

The pdf of a given assignment after conditioning on the values

BayesNets.permutecols!

BayesNets.pool

BayesNets.pool!

BayesNets.probs

BayesNets.probval

BayesNets.qqbuild

BayesNets.quantile

BayesNets.rand_bn_inference — Function.

rand_bn_inference(bn, num_query=2, num_evidence=3)

Generate a random inference state for a Bayesian Network with an evidence assignment sample uniformly over the chosen nodes' domain.

BayesNets.rand_cpd — Function.

rand_cpd(bn::DiscreteBayesNet, ncategories::Int, target::NodeName, parents::NodeNames=NodeName[])

Return a CategoricalCPD with the given number of categories with random categorical distributions

BayesNets.rand_discrete_bn — Function.

rand_discrete_bn(num_nodes16, max_num_parents=3,
        max_num_states=5, connected=true)

Generate a random DiscreteBayesNet.

Creates DiscreteBayesNet with num_nodes nodes, with each node having a random number of states and parents, up to max_num_parents and max_num_parents, respectively. If connected, each node (except the first) will be guaranteed at least one parent, making the graph connected.

BayesNets.rate

BayesNets.readtable

BayesNets.readxdsl — Function.

readxdsl( filename::AbstractString )

Return a DiscreteBayesNet read from the xdsl file

BayesNets.recode

BayesNets.recode!

BayesNets.reducedim — Function.

reducedim(op, ϕ, dims, [v0])

Reduce dimensions dims in ϕ using function op.

BayesNets.reducedim!

BayesNets.rename

BayesNets.rename!

BayesNets.required_func

BayesNets.sample

BayesNets.sample!

BayesNets.sample_posterior! — Function.

set a[varname] ~ P(varname | not varname)

Modifies a and caches in gss

BayesNets.sample_posterior_continuous! — Function.

Implements Metropolis-Hastings with a normal distribution proposal with mean equal to the previous value of the variable "varname" and stddev equal to 10 times the standard deviation of the distribution of the target variable given its parents ( var_distribution should be get(bn, varname)(a) )

MH will go through nsamples iterations. If no proposal is accepted, the original value will remain

This function expects that a[varname] is within the support of the distribution, it will not check to make sure this is true

Helper to sample_posterior Should only be used to sampling continuous distributions

set a[varname] ~ P(varname | not varname)

Modifies a and caches in gss

BayesNets.sample_posterior_finite! — Function.

Helper to sample_posterior Should only be called if the variable associated with varname is discrete

set a[varname] ~ P(varname | not varname)

Modifies both a and gss

BayesNets.sample_weighted_dataframe

BayesNets.sample_weighted_dataframe! — Function.

Chooses a sample at random from a weighted dataframe

BayesNets.sampler

BayesNets.scale

BayesNets.scale!

BayesNets.score_component — Function.

score_component(a::ScoringFunction, cpd::CPD, data::DataFrame)

Extract a Float64 score for a cpd given the data. One seeks to maximize the score.

BayesNets.score_components — Function.

score_components(a::ScoringFunction, cpd::CPD, data::DataFrame)
score_components(a::ScoringFunction, cpds::Vector{CPD}, data::DataFrame, cache::ScoreComponentCache)

Get a list of score components for all cpds

BayesNets.setindex!

BayesNets.shape

BayesNets.show

BayesNets.showall

BayesNets.showcols

BayesNets.singleton_dimension_error

BayesNets.skewness

BayesNets.span

BayesNets.sqmahal

BayesNets.sqmahal!

BayesNets.stack

BayesNets.stackdf

BayesNets.statistics — Function.

statistics(
    targetind::Int,
    parents::AbstractVector{Int},
    ncategories::AbstractVector{Int},
    data::AbstractMatrix{Int}
    )

outputs a sufficient statistics table for the target variable that is r × q where r = ncategories[i] is the number of variable instantiations and q is the number of parental instantiations of variable i

The r-values are ordered from 1 → ncategories[i] The q-values are ordered in the same ordering as ind2sub() in Julia Base Thus the instantiation of the first parent (by order given in parents[i]) is varied the fastest.

ex: Variable 1 has parents 2 and 3, with r₁ = 2, r₂ = 2, r₃ = 3 q for variable 1 is q = r₂×r₃ = 6 N will be a 6×2 matrix where: N[1,1] is the number of time v₁ = 1, v₂ = 1, v₃ = 1 N[2,1] is the number of time v₁ = 1, v₂ = 2, v₃ = 1 N[3,1] is the number of time v₁ = 1, v₂ = 1, v₃ = 2 N[4,1] is the number of time v₁ = 1, v₂ = 2, v₃ = 2 N[5,1] is the number of time v₁ = 1, v₂ = 1, v₃ = 3 N[6,1] is the number of time v₁ = 1, v₂ = 2, v₃ = 3 N[6,2] is the number of time v₁ = 2, v₂ = 1, v₃ = 1 ...

statistics(
    parent_list::Vector{Vector{Int}},
    ncategories::AbstractVector{Int},
    data::AbstractMatrix{Int},
    )

Computes sufficient statistics from a discrete dataset for a Discrete Bayesian Net structure

INPUT: parents: list of lists of parent indices A variable with index i has ncategories[i] and row in data[i,:] No acyclicity checking is done ncategories: list of variable bin counts, or number of discrete values the variable can take on, v ∈ {1 : ncategories[i]} data: table of discrete values [n×m] where n is the number of nodes and m is the number of samples

OUTPUT: N :: Vector{Matrix{Int}} a sufficient statistics table for each variable Variable with index i has statistics table N[i], which is r × q where r = ncategories[i] is the number of variable instantiations and q is the number of parental instantiations of variable i

    The r-values are ordered from 1 → ncategories[i]
    The q-values are ordered in the same ordering as ind2sub() in Julia Base
        Thus the instantiation of the first parent (by order given in parents[i])
        is varied the fastest.

    ex:
        Variable 1 has parents 2 and 3, with r₁ = 2, r₂ = 2, r₃ = 3
        q for variable 1 is q = r₂×r₃ = 6
        N[1] will be a 6×2 matrix where:
            N[1][1,1] is the number of time v₁ = 1, v₂ = 1, v₃ = 1
            N[1][2,1] is the number of time v₁ = 1, v₂ = 2, v₃ = 1
            N[1][3,1] is the number of time v₁ = 1, v₂ = 1, v₃ = 2
            N[1][4,1] is the number of time v₁ = 1, v₂ = 2, v₃ = 2
            N[1][5,1] is the number of time v₁ = 1, v₂ = 1, v₃ = 3
            N[1][6,1] is the number of time v₁ = 1, v₂ = 2, v₃ = 3
            N[1][6,2] is the number of time v₁ = 2, v₂ = 1, v₃ = 1
            ...

This function uses sparse matrix black magic and was mercilessly stolen from Ed Schmerling.

BayesNets.std

BayesNets.stdlogx

BayesNets.CPDs.strip_arg — Function.

strip_arg(arg::Symbol)

Strip anything extra (type annotations, default values, etc) from an argument. For now this cannot handle keyword arguments (it will throw an error).

BayesNets.sub

BayesNets.sub2ind_vec

BayesNets.succprob

BayesNets.suffstats

BayesNets.sumout — Function.

sumout(t, v)

Table marginalization

BayesNets.support

BayesNets.table — Function.

table(bn::DiscreteBayesNet, name::NodeName)

Constructs the CPD factor associated with the given node in the BayesNet

BayesNets.tail

BayesNets.test_distr

BayesNets.test_samples

BayesNets.uncompact

BayesNets.unique!

BayesNets.unstack

BayesNets.values

BayesNets.var

BayesNets.varlogx

BayesNets.vcat

BayesNets.writetable

BayesNets.wsample

BayesNets.wsample!