Docstrings

CrossfilterCharts.AbstractBubbleChart

CrossfilterCharts.Attribute

CrossfilterCharts.BarChart

CrossfilterCharts.BaseChart

CrossfilterCharts.BubbleChart

CrossfilterCharts.ChartType

CrossfilterCharts.ColorChart

CrossfilterCharts.CoordinateGridChart

CrossfilterCharts.CrossfilterCharts

CrossfilterCharts.DCChart

CrossfilterCharts.DCOut

CrossfilterCharts.DCWidget

CrossfilterCharts.DataCountWidget

CrossfilterCharts.DataTableWidget

CrossfilterCharts.Dependency

CrossfilterCharts.Dimension

CrossfilterCharts.Group

CrossfilterCharts.LineChart

CrossfilterCharts.NULL_ATTRIBUTE

CrossfilterCharts.PieChart

CrossfilterCharts.RowChart

CrossfilterCharts.WidgetChart

CrossfilterCharts._generate_accessor

CrossfilterCharts.add_bubblechart! — Function.

add_bubblechart!

Construct and append a Bubble Chart to the DCOut object. Requires a previously created dimension. xcol, ycol, and r_col denote the DataFrame fields whose sums will determine the x position, y position, and radius of the bubbles in the final chart.

Use :DCCount to access the count field.

CrossfilterCharts.add_chart! — Function.

add_chart!

Append the DCChart to the list of charts in the DCOut object.

CrossfilterCharts.add_datacountwidget! — Function.

add_datacountwidget!

Construct and append a Data Count Widget to the DCOut object.

CrossfilterCharts.add_datatablewidget! — Function.

add_datatablewidget!

Construct and append a Data Table Widget to the DCOut object. Requires a previously constructed dimension for the first column.

CrossfilterCharts.add_dimension! — Function.

add_dimension!

Append the Dimension to the list of dimensions in the DCOut object.

CrossfilterCharts.add_group! — Function.

add_group!

Append the Group to the list of groups in the DCOut object.

CrossfilterCharts.add_widget! — Function.

add_widget!

Append the DCWidget to the list of widgets in the DCOut object.

CrossfilterCharts.barchart — Function.

barchart

Infer construction of a DC barchart based on the given group.

CrossfilterCharts.bubblechart — Function.

bubblechart

Construct an empty custom DC bubblechart.

bubbleChart

Construct a bubblechart using the given master grouping and sums given by x_col, y_col, and r_col for x position, y position, and radius.

CrossfilterCharts.can_infer_chart

CrossfilterCharts.clear_charts! — Function.

clear_charts!

Remove all charts from the DCOut object.

CrossfilterCharts.clear_widgets! — Function.

clear_widgets

Remove all widgets from the DCOut object.

CrossfilterCharts.datacountwidget — Function.

datacountwidget

Construct a DC DataCountWidget.

CrossfilterCharts.datatablewidget — Function.

datatablewidget

Construct a DC DataTableWidget.

CrossfilterCharts.dc — Function.

dc(df::DataFrame)

Construct a DC.js visualization based on the columns in the given DataFrame. This is the easiest and most straightforward way to use DC.jl.

CrossfilterCharts.eval

CrossfilterCharts.get_all_attributes

CrossfilterCharts.get_charts — Function.

get_charts

Returns all charts constructed from the given column of the given type. chart_type can be: piechart, barchart, linechart, rowchart, bubblechart

CrossfilterCharts.get_dim_by_col — Function.

get_group_by_name

Returns the dimension inside the given DCCout instance created from the given column.

CrossfilterCharts.get_group_by_name — Function.

get_group_by_name

Returns the group inside the given DCCout instance with the given name.

CrossfilterCharts.get_groups_by_col — Function.

get_groups_by_col

Returns the groups inside the given DCCout instance associated with the given column.

CrossfilterCharts.include

CrossfilterCharts.infer_chart — Function.

infer_chart(arr::AbstractVector, group::Group)

Constructs a Chart suitable for the type in arr.

CrossfilterCharts.infer_dimension — Function.

infer_dimension

Constructs a Dimension suitable for the type in arr.

CrossfilterCharts.infer_dimensions! — Function.

infer_dimensions!

Infer a dimension for each column in the DCOut DataFrame

CrossfilterCharts.infer_group — Function.

infer_group

Infer construction of a group based on the array datatype.

CrossfilterCharts.infer_groups! — Function.

infer_groups!

Infer a group for each dimension in the DCOut DataFrame

CrossfilterCharts.isna — Function.

can_infer_chart(arr::AbstractVector)

Whether chart inference is supported for the given array type.

CrossfilterCharts.linechart — Function.

linechart

Infer construction of a DC linechart based on the given group.

CrossfilterCharts.piechart — Function.

piechart

Infer construction of a DC piechart based on the given group.

CrossfilterCharts.quick_add! — Function.

quick_add!

A utility function for quickly building a chart and adding it. Works with linechart, barchart, rowchart, piechart. Requires previously a constructed group.

quick_add!

Quickly build a chart using the group constructed from column. If multiple groups are found, an error is thrown unless use_first is set, in which case the first group found is used.

quick_add!

Quickly build a chart using the group with the given name.

CrossfilterCharts.quote_corrector

CrossfilterCharts.randomize_ids! — Function.

randomize_ids!

Randomly re-initialize all dcout random ids. These are used when exporting to HTML to prevent charts from referencing one another across IJulia cells.

CrossfilterCharts.randomize_parent

CrossfilterCharts.reduce_count — Function.

reduce_count

A reduction for a Group that is simply reduceCount()

CrossfilterCharts.reduce_master — Function.

reduce_sum

A master reduction which sums values from all provided columns and tallies a count. Useful for making more complex charts like bubble charts.

CrossfilterCharts.reduce_sum — Function.

reduce_sum

A reduction for a Group that sums the values.

CrossfilterCharts.remove_chart! — Function.

remove_chart!

Removes the given DCChart from the list of charts in the DCOut object.

CrossfilterCharts.remove_group! — Function.

remove_group!

Removes the given Group from the list of groups in the DCOut object.

CrossfilterCharts.round_to_nearest_half_order_of_magnitude — Function.

round_to_nearest_half_order_of_magnitude

Rounds a number to the nearest half order of magnitude, {...0.1,0.5,1,5,10,50,100,500...}

CrossfilterCharts.rowchart — Function.

rowchart

Infer construction of a DC rowchart based on the given group.

CrossfilterCharts.scale_default

CrossfilterCharts.set_elastic_height! — Function.

set_elastic_height!

Changes the elasticity of the output div height. A value of true indicates that the div will resize to fit all elements without needing a scrollbar.

CrossfilterCharts.size_default!

CrossfilterCharts.write_data

CrossfilterCharts.write_html_body

CrossfilterCharts.write_html_chart_entry

CrossfilterCharts.write_html_head

CrossfilterCharts.write_html_widget_entry

CrossfilterCharts.write_json_entry

CrossfilterCharts.write_reset_script

CrossfilterCharts.write_script

CrossfilterCharts.write_script_dependencies

CrossfilterCharts.write_source_html