Docstrings

GLFW.@callback

GLFW.@checked_lib

GLFW.@windowcallback

GLFW.ACCUM_ALPHA_BITS

GLFW.ACCUM_BLUE_BITS

GLFW.ACCUM_GREEN_BITS

GLFW.ACCUM_RED_BITS

GLFW.ALPHA_BITS

GLFW.ANY_RELEASE_BEHAVIOR

GLFW.API_UNAVAILABLE

GLFW.ARROW_CURSOR

GLFW.AUTO_ICONIFY

GLFW.AUX_BUFFERS

GLFW.Action

GLFW.BLUE_BITS

GLFW.CLIENT_API

GLFW.CONNECTED

GLFW.CONTEXT_CREATION_API

GLFW.CONTEXT_NO_ERROR

GLFW.CONTEXT_RELEASE_BEHAVIOR

GLFW.CONTEXT_REVISION

GLFW.CONTEXT_ROBUSTNESS

GLFW.CONTEXT_VERSION_MAJOR

GLFW.CONTEXT_VERSION_MINOR

GLFW.CROSSHAIR_CURSOR

GLFW.CURSOR

GLFW.CURSOR_DISABLED

GLFW.CURSOR_HIDDEN

GLFW.CURSOR_NORMAL

GLFW.Callback

GLFW.CreateStandardCursor

GLFW.CreateWindow

GLFW.CreateWindowSurface — Function.

CreateWindowSurface(instance, window, allocator=C_NULL)

Create a Vulkan surface for the specified window.

GLFW.Cursor

GLFW.DECORATED

GLFW.DEPTH_BITS

GLFW.DISCONNECTED

GLFW.DONT_CARE

GLFW.DOUBLEBUFFER

GLFW.DefaultWindowHints

GLFW.DestroyCursor

GLFW.DestroyWindow

GLFW.DeviceConfigEvent

GLFW.EGL_CONTEXT_API

GLFW.ErrorCode

GLFW.ExtensionSupported

GLFW.FLOATING

GLFW.FOCUSED

GLFW.FORMAT_UNAVAILABLE

GLFW.GLFW

GLFW.GLFWError

GLFW.GLFWImage

GLFW.GREEN_BITS

GLFW.GetCurrentContext

GLFW.GetCursorPos

GLFW.GetFramebufferSize

GLFW.GetInputMode

GLFW.GetInstanceProcAddress — Function.

GetInstanceProcAddress(instance, procname) -> funcptr

Return the address of the specified Vulkan core or extension function for the specified instance. funcptr can be used directly as the first argument of ccall: ccall(funcptr, ...).

GLFW.GetJoystickAxes

GLFW.GetJoystickButtons

GLFW.GetJoystickName

GLFW.GetKey

GLFW.GetKeyName

GLFW.GetMonitorName

GLFW.GetMonitorPhysicalSize

GLFW.GetMonitorPos

GLFW.GetMonitors

GLFW.GetMouseButton

GLFW.GetPhysicalDevicePresentationSupport — Function.

GetPhysicalDevicePresentationSupport(instance, device, queuefamily)

Return whether the specified queue family of the specified physical device supports presentation to the platform GLFW was built for.

GLFW.GetPrimaryMonitor

GLFW.GetProcAddress

GLFW.GetRequiredInstanceExtensions — Function.

GetRequiredInstanceExtensions()

Return a vector of names of Vulkan instance extensions.

GLFW.GetVersion

GLFW.GetVersionString

GLFW.GetVideoMode

GLFW.GetVideoModes

GLFW.GetWindowAttrib

GLFW.GetWindowFrameSize

GLFW.GetWindowMonitor

GLFW.GetWindowPos

GLFW.GetWindowSize

GLFW.HAND_CURSOR

GLFW.HRESIZE_CURSOR

GLFW.HideWindow

GLFW.IBEAM_CURSOR

GLFW.ICONIFIED

GLFW.INITIALIZED

GLFW.INVALID_ENUM

GLFW.INVALID_VALUE

GLFW.IconifyWindow

GLFW.Init

GLFW.JOYSTICK_1

GLFW.JOYSTICK_10

GLFW.JOYSTICK_11

GLFW.JOYSTICK_12

GLFW.JOYSTICK_13

GLFW.JOYSTICK_14

GLFW.JOYSTICK_15

GLFW.JOYSTICK_16

GLFW.JOYSTICK_2

GLFW.JOYSTICK_3

GLFW.JOYSTICK_4

GLFW.JOYSTICK_5

GLFW.JOYSTICK_6

GLFW.JOYSTICK_7

GLFW.JOYSTICK_8

GLFW.JOYSTICK_9

GLFW.Joystick

GLFW.JoystickPresent

GLFW.KEY_0

GLFW.KEY_1

GLFW.KEY_2

GLFW.KEY_3

GLFW.KEY_4

GLFW.KEY_5

GLFW.KEY_6

GLFW.KEY_7

GLFW.KEY_8

GLFW.KEY_9

GLFW.KEY_A

GLFW.KEY_APOSTROPHE

GLFW.KEY_B

GLFW.KEY_BACKSLASH

GLFW.KEY_BACKSPACE

GLFW.KEY_C

GLFW.KEY_CAPS_LOCK

GLFW.KEY_COMMA

GLFW.KEY_D

GLFW.KEY_DELETE

GLFW.KEY_DOWN

GLFW.KEY_E

GLFW.KEY_END

GLFW.KEY_ENTER

GLFW.KEY_EQUAL

GLFW.KEY_ESCAPE

GLFW.KEY_F

GLFW.KEY_F1

GLFW.KEY_F10

GLFW.KEY_F11

GLFW.KEY_F12

GLFW.KEY_F13

GLFW.KEY_F14

GLFW.KEY_F15

GLFW.KEY_F16

GLFW.KEY_F17

GLFW.KEY_F18

GLFW.KEY_F19

GLFW.KEY_F2

GLFW.KEY_F20

GLFW.KEY_F21

GLFW.KEY_F22

GLFW.KEY_F23

GLFW.KEY_F24

GLFW.KEY_F25

GLFW.KEY_F3

GLFW.KEY_F4

GLFW.KEY_F5

GLFW.KEY_F6

GLFW.KEY_F7

GLFW.KEY_F8

GLFW.KEY_F9

GLFW.KEY_G

GLFW.KEY_GRAVE_ACCENT

GLFW.KEY_H

GLFW.KEY_HOME

GLFW.KEY_I

GLFW.KEY_INSERT

GLFW.KEY_J

GLFW.KEY_K

GLFW.KEY_KP_0

GLFW.KEY_KP_1

GLFW.KEY_KP_2

GLFW.KEY_KP_3

GLFW.KEY_KP_4

GLFW.KEY_KP_5

GLFW.KEY_KP_6

GLFW.KEY_KP_7

GLFW.KEY_KP_8

GLFW.KEY_KP_9

GLFW.KEY_KP_ADD

GLFW.KEY_KP_DECIMAL

GLFW.KEY_KP_DIVIDE

GLFW.KEY_KP_ENTER

GLFW.KEY_KP_EQUAL

GLFW.KEY_KP_MULTIPLY

GLFW.KEY_KP_SUBTRACT

GLFW.KEY_L

GLFW.KEY_LEFT

GLFW.KEY_LEFT_ALT

GLFW.KEY_LEFT_BRACKET

GLFW.KEY_LEFT_CONTROL

GLFW.KEY_LEFT_SHIFT

GLFW.KEY_LEFT_SUPER

GLFW.KEY_M

GLFW.KEY_MENU

GLFW.KEY_MINUS

GLFW.KEY_N

GLFW.KEY_NUM_LOCK

GLFW.KEY_O

GLFW.KEY_P

GLFW.KEY_PAGE_DOWN

GLFW.KEY_PAGE_UP

GLFW.KEY_PAUSE

GLFW.KEY_PERIOD

GLFW.KEY_PRINT_SCREEN

GLFW.KEY_Q

GLFW.KEY_R

GLFW.KEY_RIGHT

GLFW.KEY_RIGHT_ALT

GLFW.KEY_RIGHT_BRACKET

GLFW.KEY_RIGHT_CONTROL

GLFW.KEY_RIGHT_SHIFT

GLFW.KEY_RIGHT_SUPER

GLFW.KEY_S

GLFW.KEY_SCROLL_LOCK

GLFW.KEY_SEMICOLON

GLFW.KEY_SLASH

GLFW.KEY_SPACE

GLFW.KEY_T

GLFW.KEY_TAB

GLFW.KEY_U

GLFW.KEY_UNKNOWN

GLFW.KEY_UP

GLFW.KEY_V

GLFW.KEY_W

GLFW.KEY_WORLD_1

GLFW.KEY_WORLD_2

GLFW.KEY_X

GLFW.KEY_Y

GLFW.KEY_Z

GLFW.Key

GLFW.LOSE_CONTEXT_ON_RESET

GLFW.MAXIMIZED

GLFW.MOD_ALT

GLFW.MOD_CONTROL

GLFW.MOD_SHIFT

GLFW.MOD_SUPER

GLFW.MOUSE_BUTTON_1

GLFW.MOUSE_BUTTON_2

GLFW.MOUSE_BUTTON_3

GLFW.MOUSE_BUTTON_4

GLFW.MOUSE_BUTTON_5

GLFW.MOUSE_BUTTON_6

GLFW.MOUSE_BUTTON_7

GLFW.MOUSE_BUTTON_8

GLFW.MOUSE_BUTTON_LEFT

GLFW.MOUSE_BUTTON_MIDDLE

GLFW.MOUSE_BUTTON_RIGHT

GLFW.MakeContextCurrent

GLFW.MaximizeWindow

GLFW.Monitor

GLFW.MouseButton

GLFW.NATIVE_CONTEXT_API

GLFW.NOT_INITIALIZED

GLFW.NO_API

GLFW.NO_CURRENT_CONTEXT

GLFW.NO_RESET_NOTIFICATION

GLFW.NO_ROBUSTNESS

GLFW.NO_WINDOW_CONTEXT

GLFW.OPENGL_ANY_PROFILE

GLFW.OPENGL_API

GLFW.OPENGL_COMPAT_PROFILE

GLFW.OPENGL_CORE_PROFILE

GLFW.OPENGL_DEBUG_CONTEXT

GLFW.OPENGL_ES_API

GLFW.OPENGL_FORWARD_COMPAT

GLFW.OPENGL_PROFILE

GLFW.OUT_OF_MEMORY

GLFW.PLATFORM_ERROR

GLFW.PRESS

GLFW.PollEvents

GLFW.PostEmptyEvent

GLFW.RED_BITS

GLFW.REFRESH_RATE

GLFW.RELEASE

GLFW.RELEASE_BEHAVIOR_FLUSH

GLFW.RELEASE_BEHAVIOR_NONE

GLFW.REPEAT

GLFW.RESIZABLE

GLFW.RestoreWindow

GLFW.SAMPLES

GLFW.SRGB_CAPABLE

GLFW.STENCIL_BITS

GLFW.STEREO

GLFW.STICKY_KEYS

GLFW.STICKY_MOUSE_BUTTONS

GLFW.SetCharCallback

GLFW.SetCharModsCallback

GLFW.SetCursor

GLFW.SetCursorEnterCallback

GLFW.SetCursorPos

GLFW.SetCursorPosCallback

GLFW.SetDropCallback

GLFW.SetErrorCallback

GLFW.SetFramebufferSizeCallback

GLFW.SetGamma

GLFW.SetInputMode

GLFW.SetJoystickCallback

GLFW.SetKeyCallback

GLFW.SetMonitorCallback

GLFW.SetMouseButtonCallback

GLFW.SetScrollCallback

GLFW.SetWindowAspectRatio

GLFW.SetWindowCloseCallback

GLFW.SetWindowFocusCallback

GLFW.SetWindowIcon — Function.

GLFW.SetWindowIcon(window::Window, image::Matrix{NTuple{4, UInt8}})

GLFW.SetWindowIcon(window::Window, images::Vector{<:AbstractMatrix{NTuple{4,UInt8}}})

Set the window icon, where a single image may be passed or a vector of images with different icon sizes. The images must be of RGBA format. Before calling this function it might be necessary to reinterpret the image as a matrix of element type NTuple{4, UInt8} if the icons are loaded with type RGBA{N0f8}

Examples

using FileIO
icons = load.(["icon-16.png", "icon-32.png", "icon-128.png"])
buffs = reinterpret.(NTuple{4, UInt8}, icons)
GLFW.SetWindowIcon(win, buffs)
GLFW.PollEvents() # needs a poll events to become active

GLFW.SetWindowIconifyCallback

GLFW.SetWindowMonitor

GLFW.SetWindowPos

GLFW.SetWindowPosCallback

GLFW.SetWindowRefreshCallback

GLFW.SetWindowShouldClose

GLFW.SetWindowSize

GLFW.SetWindowSizeCallback

GLFW.SetWindowSizeLimits

GLFW.SetWindowTitle

GLFW.ShowWindow

GLFW.StandardCursorShape

GLFW.SwapBuffers

GLFW.SwapInterval

GLFW.Terminate

GLFW.VERSION_UNAVAILABLE

GLFW.VISIBLE

GLFW.VRESIZE_CURSOR

GLFW.VidMode

GLFW.VkAllocationCallbacks

GLFW.VkInstance

GLFW.VkPhysicalDevice

GLFW.VkResult

GLFW.VkSurfaceKHR

GLFW.VulkanSupported — Function.

VulkanSupported()

Return whether the Vulkan loader has been found.

GLFW.WaitEvents

GLFW.Window — Type.

Function to create a pure GLFW OpenGL window

GLFW.WindowHint

GLFW.WindowShouldClose

GLFW._CharCallbackWrapper

GLFW._CharModsCallbackWrapper

GLFW._CursorEnterCallbackWrapper

GLFW._CursorPosCallbackWrapper

GLFW._DropCallbackWrapper

GLFW._ErrorCallbackWrapper

GLFW._FramebufferSizeCallbackWrapper

GLFW._JoystickCallbackWrapper

GLFW._KeyCallbackWrapper

GLFW._MonitorCallbackWrapper

GLFW._MouseButtonCallbackWrapper

GLFW._ScrollCallbackWrapper

GLFW._WindowCloseCallbackWrapper

GLFW._WindowFocusCallbackWrapper

GLFW._WindowIconifyCallbackWrapper

GLFW._WindowPosCallbackWrapper

GLFW._WindowRefreshCallbackWrapper

GLFW._WindowSizeCallbackWrapper

GLFW.__init__

GLFW._window_callback_num

GLFW._window_callbacks

GLFW.arrowsplit

GLFW.callbackcode

GLFW.callbacks

GLFW.depsjl_path

GLFW.eval

GLFW.extractargs

GLFW.include

GLFW.is_initialized

GLFW.lib

GLFW.make_fullscreen!

GLFW.make_windowed!

GLFW.paramname

GLFW.paramtype

GLFW.primarymonitorresolution — Function.

Returns the monitor resolution of the primary monitor.

GLFW.set_visibility! — Function.

Sets visibility of OpenGL window. Will still render if not visible. Only applies to the root screen holding the opengl context.

GLFW.standard_context_hints — Function.

Tries to create sensible context hints! Taken from lessons learned at: GLFW

GLFW.standard_screen_resolution — Function.

Takes half the resolution of the primary monitor. This should make for sensible defaults!

GLFW.standard_window_hints — Function.

Standard window hints for creating a plain context without any multisampling or extra buffers beside the color buffer