Docstrings

SauterSchwabQuadrature.CommonEdge

SauterSchwabQuadrature.CommonFace

SauterSchwabQuadrature.CommonVertex

SauterSchwabQuadrature.PositiveDistance

SauterSchwabQuadrature.SauterSchwabQuadrature

SauterSchwabQuadrature.SauterSchwabStrategy

SauterSchwabQuadrature._legendre

SauterSchwabQuadrature.eval

SauterSchwabQuadrature.generate_integrand_uv — Function.

Generates the integrand for use in the Sauter-Schwab quadrature formulas.

    generate_integrand(kernel, test_local_space, trial_local_space,
            test_chart, trial_chart)

kernel is a function that takes two neighborhoods x and y and returns a scalar or dyadic value. kernel is the part of the integrand that is most easily described in term of cartesian coordinates, that is non-separable in the two arguments, and that does not depend on which of the local shape functions is considered. In boundary element methods it is the Green function (fundamental solution) of the underlying partial differential equations.

SauterSchwabQuadrature.include

SauterSchwabQuadrature.k3p_ce

SauterSchwabQuadrature.k3p_cf

SauterSchwabQuadrature.k3p_cv

SauterSchwabQuadrature.k3p_pd

SauterSchwabQuadrature.reorder

SauterSchwabQuadrature.sauterschwab_parameterized — Function.

Compute interaction integrals using the quadrature introduced in [1].

sauterschwab_parameterized(integrand, strategy)

Here, integrand is the pull-back of the integrand into the barycentric parameter domain of the two triangles that define the integration domain. Note that here we use (for a planar triangle) the representation:

x = x[3] + u*(x[1]-x[3]) + v*(x[2]-x[3])

with u ranging from 0 to 1 and v ranging from 0 to 1-u. This parameter domain and representation is different from the one used in [1].

The second argument 'strategy' is an object whose type is one of

CommonFace

CommonEdge

CommonVertex

PositiveDistance

according to the configuration of the two triangular patches defining the domain of integration. The constructors of these classes take a single argument acc that defines the number of quadrature points along each of the four axes of the final rectangular (ξ,η) integration domain (see [1], Ch 5).

[1] Sauter. Schwwab, 'Boundary Element Methods', Springer Berlin Heidelberg, 2011