Docstrings

Neo4j.Connection — Type.

Connection()

Examples

julia> con = Neo4j.Connection("localhost")
Neo4j.Connection(false, "localhost", 7474, "/db/data/", "http://localhost:7474/db/data/", "", "")

Neo4j.DEFAULT_HOST

Neo4j.DEFAULT_PORT

Neo4j.DEFAULT_URI

Neo4j.Graph

Neo4j.JSONArray

Neo4j.JSONData

Neo4j.JSONObject

Neo4j.Neo4j

Neo4j.Node

Neo4j.QueryData

Neo4j.Relationship

Neo4j.Result

Neo4j.Statement

Neo4j.Transaction

Neo4j.addnodelabel

Neo4j.addnodelabels

Neo4j.commit

Neo4j.connheaders

Neo4j.connurl

Neo4j.createnode

Neo4j.createrel

Neo4j.cypherQuery — Function.

cypherQuery(conn, cypher, params; elTypes, nRowsElTypeCheck)

Retrieve molecular identifier from other databases, targetDb, for single or mulitple query IDs, queryId, and moreover information on Ensembl gene, transcript and peptide IDs, such as ID and genomic loation.

Arguments

conn::Neo4j.Connection : a valid connection to a Neo4j graph DB instance.

cypher::String : Cypher MATCH query returning tabular data.

params::Pair : parameters which are passed on to the cypher query.

elTypes::Vector{Type} : column types can be provided manually as a Vector{Type}

nRowsElTypeCheck::Int : Number of rows which are used to determine column datatypes (defaults to 1000)

Examples

julia> cypherQuery(
         Neo4j.Connection("localhost"),
         "MATCH (p :Person {name: {name}}) RETURN p.name AS Name, p.age AS Age;",
         "name" => "John Doe")

Neo4j.deletenode

Neo4j.deletenodelabel

Neo4j.deletenodeproperties

Neo4j.deletenodeproperty

Neo4j.deleterel

Neo4j.eval

Neo4j.getElTypes

Neo4j.getgraph

Neo4j.getlabels

Neo4j.getneighbors

Neo4j.getnode

Neo4j.getnodelabels

Neo4j.getnodeproperties

Neo4j.getnodeproperty

Neo4j.getnodesforlabel

Neo4j.getrel

Neo4j.getrelproperties

Neo4j.getrelproperty

Neo4j.getrels

Neo4j.include

Neo4j.parseResults

Neo4j.request

Neo4j.rollback

Neo4j.setnodeproperty

Neo4j.transaction

Neo4j.updatenodelabels

Neo4j.updatenodeproperties

Neo4j.updaterelproperties

Neo4j.version