Docstrings

LibSndFile.@checked_lib

LibSndFile.LengthIO

LibSndFile.LibSndFile

LibSndFile.SFM_READ

LibSndFile.SFM_WRITE

LibSndFile.SF_ENDIAN_BIG

LibSndFile.SF_ENDIAN_CPU

LibSndFile.SF_ENDIAN_FILE

LibSndFile.SF_ENDIAN_LITTLE

LibSndFile.SF_FORMAT_AIFF

LibSndFile.SF_FORMAT_ALAW

LibSndFile.SF_FORMAT_AU

LibSndFile.SF_FORMAT_AVR

LibSndFile.SF_FORMAT_CAF

LibSndFile.SF_FORMAT_DOUBLE

LibSndFile.SF_FORMAT_DPCM_16

LibSndFile.SF_FORMAT_DPCM_8

LibSndFile.SF_FORMAT_DWVW_12

LibSndFile.SF_FORMAT_DWVW_16

LibSndFile.SF_FORMAT_DWVW_24

LibSndFile.SF_FORMAT_DWVW_N

LibSndFile.SF_FORMAT_ENDMASK

LibSndFile.SF_FORMAT_FLAC

LibSndFile.SF_FORMAT_FLOAT

LibSndFile.SF_FORMAT_G721_32

LibSndFile.SF_FORMAT_G723_24

LibSndFile.SF_FORMAT_G723_40

LibSndFile.SF_FORMAT_GSM610

LibSndFile.SF_FORMAT_HTK

LibSndFile.SF_FORMAT_IMA_ADPCM

LibSndFile.SF_FORMAT_IRCAM

LibSndFile.SF_FORMAT_MAT4

LibSndFile.SF_FORMAT_MAT5

LibSndFile.SF_FORMAT_MPC2K

LibSndFile.SF_FORMAT_MS_ADPCM

LibSndFile.SF_FORMAT_NIST

LibSndFile.SF_FORMAT_OGG

LibSndFile.SF_FORMAT_PAF

LibSndFile.SF_FORMAT_PCM_16

LibSndFile.SF_FORMAT_PCM_24

LibSndFile.SF_FORMAT_PCM_32

LibSndFile.SF_FORMAT_PCM_S8

LibSndFile.SF_FORMAT_PCM_U8

LibSndFile.SF_FORMAT_PVF

LibSndFile.SF_FORMAT_RAW

LibSndFile.SF_FORMAT_RF64

LibSndFile.SF_FORMAT_SD2

LibSndFile.SF_FORMAT_SDS

LibSndFile.SF_FORMAT_SUBMASK

LibSndFile.SF_FORMAT_SVX

LibSndFile.SF_FORMAT_TYPEMASK

LibSndFile.SF_FORMAT_ULAW

LibSndFile.SF_FORMAT_VOC

LibSndFile.SF_FORMAT_VORBIS

LibSndFile.SF_FORMAT_VOX_ADPCM

LibSndFile.SF_FORMAT_W64

LibSndFile.SF_FORMAT_WAV

LibSndFile.SF_FORMAT_WAVEX

LibSndFile.SF_FORMAT_WVE

LibSndFile.SF_FORMAT_XI

LibSndFile.SF_INFO

LibSndFile.SF_SEEK_CUR

LibSndFile.SF_SEEK_END

LibSndFile.SF_SEEK_SET

LibSndFile.SF_VIRTUAL_IO

LibSndFile.SndFileSink

LibSndFile.SndFileSource

LibSndFile.SndFileStream

LibSndFile.__init__

LibSndFile.depsjl

LibSndFile.eval

LibSndFile.fmt_to_type — Function.

Take a LibSndFile format code and return a suitable sample type

LibSndFile.formatcode

LibSndFile.include

LibSndFile.inferlen — Function.

inferlen(io)

Try to infer the length of io in bytes

LibSndFile.libsndfile

LibSndFile.load

LibSndFile.load_helper

LibSndFile.loadstreaming

LibSndFile.save

LibSndFile.save_helper

LibSndFile.savestreaming

LibSndFile.sf_close

LibSndFile.sf_count_t

LibSndFile.sf_open

LibSndFile.sf_readf — Function.

Wrappers for the family of sfreadf* functions, which read the given number of frames into the given array. Returns the number of frames read.

LibSndFile.sf_strerror

LibSndFile.sf_writef — Function.

Wrappers for the family of sfwritef* functions, which write the given number of frames into the given array. Returns the number of frames written.

LibSndFile.subformatcode

LibSndFile.supported_formats

LibSndFile.virtual_get_filelen

LibSndFile.virtual_read

LibSndFile.virtual_seek

LibSndFile.virtual_tell

LibSndFile.virtual_write