Docstrings

SampledSignals...

SampledSignals.AbstractSampleBuf

SampledSignals.ArrayIsh

SampledSignals.BuiltinIdx

SampledSignals.BuiltinMultiIdx

SampledSignals.ClosedInterval

SampledSignals.ConvertIdx

SampledSignals.DEFAULT_BLOCKSIZE

SampledSignals.DownMixSink — Type.

DownMixSink provides a multi-channel sink that wraps a single-channel sink. Writing to this sink mixes all the channels down to the single channel

SampledSignals.FrameQuant

SampledSignals.GHz

SampledSignals.Hz

SampledSignals.MHz

SampledSignals.PCM16Sample

SampledSignals.PCM20Sample

SampledSignals.PCM24Sample

SampledSignals.PCM32Sample

SampledSignals.PCM64Sample

SampledSignals.PCM8Sample

SampledSignals.ReformatSink

SampledSignals.ResampleSink

SampledSignals.SAMPLE_TYPE

SampledSignals.SampleBuf — Type.

Represents a multi-channel regularly-sampled buffer that stores its own sample rate (in samples/second). The wrapped data is an N-dimensional array. A 1-channel sample can be represented with a 1D array or an Mx1 matrix, and a C-channel buffer will be an MxC matrix. So a 1-second stereo audio buffer sampled at 44100Hz with 32-bit floating-point samples in the time domain would have the type SampleBuf{Float32, 2}.

SampledSignals.SampleBufSink — Type.

SampleBufSink is a SampleSink backed by a buffer. It's mostly useful to hook into the stream conversion infrastructure, because you can wrap a buffer in a SampleBufSink and then read a source into it with a different channel count, sample rate, or channel count.

SampledSignals.SampleBufSource — Type.

SampleBufSource is a SampleSource backed by a buffer. It's mostly useful to hook into the stream conversion infrastructure, because you can wrap a buffer in a SampleBufSource and then write it into a sink with a different channel count, sample rate, or channel count.

SampledSignals.SampleRate

SampledSignals.SampleSink — Type.

Represents a sink that samples can be written to, such as an audio file or headphone output.

Subtypes should implement the samplerate, nchannels, eltype, and unsafe_write methods. unsafe_write can assume that the samplerate, channel count, and element type are all matching.

SampledSignals.SampleSource — Type.

Represents a source of samples, such as an audio file, microphone input, or SDR Receiver.

Subtypes should implement the samplerate, nchannels, eltype, and unsafe_read! methods. unsafe_read! can assume that the samplerate, channel count, and element type are all matching.

SampledSignals.SampledSignals

SampledSignals.SinSource — Type.

SinSource(eltype, samplerate, freqs)

SinSource is a multi-channel sine-tone signal generator. freqs can be an array of frequencies for a multi-channel source, or a single frequency for a mono source.

SampledSignals.SpectrumBuf — Type.

Represents a multi-channel regularly-sampled buffer representing the frequency- domain spectrum of a SampleBuf. The wrapped data is an N-dimensional array. A 1-channel sample can be represented with a 1D array or an Mx1 matrix, and a C-channel buffer will be an MxC matrix. So a 1-second stereo audio buffer sampled at 44100Hz with 32-bit floating-point samples in the time domain would have the type SampleBuf{Float32, 2}.

SampledSignals.THz

SampledSignals.UpMixSink — Type.

UpMixSink provides a single-channel sink that wraps a multi-channel sink. Writing to this sink copies the single channel to all the channels in the wrapped sink

SampledSignals.WAVE_FORMAT_PCM

SampledSignals.WAVFormat

SampledSignals.__init__

SampledSignals.blocksize

SampledSignals.channelptr — Function.

Get a pointer to the underlying data for the buffer. Will return a Ptr{T}, where T is the element type of the buffer. This is particularly useful for passing to C libraries to fill the buffer

SampledSignals.dB

SampledSignals.domain

SampledSignals.embed_javascript

SampledSignals.eval

SampledSignals.find_buf

SampledSignals.frames

SampledSignals.inHz — Function.

inHz(quantity[, rate])

Translate a particular quantity (usually a frequency) to a (unitless) value in Hz.

If the given quantity is Unitful, we use the given units. If it is not we assume it is already a value in Hz.

For some units (e.g. frames) you will need to specify a sample rate:

Example

julia> inHz(1.0kHz) 1000.0

SampledSignals.include

SampledSignals.inframes — Function.

inframes([Type,]quantity[, rate])

Translate the given quantity to a (unitless) number of time or frequency frames, given a particular samplerate. Note that this isn't quantized to integer numbers of frames. If given a Type, the result will first be coerced to the given type.

If the given quantity is Unitful, we use the given units. If it is not we assume it is already a value in frames.

Example

julia> inframes(0.5s, 44100Hz) 22050.0

julia> inframes(1000Hz, 2048/44100Hz) 46.439909297052154

SampledSignals.inseconds — Function.

inseconds(quantity[, rate])

Translate a particular quantity (usually a time) to a (unitless) value in seconds.

If the given quantity is Unitful, we use the given units. If it is not we assume it is already a value in seconds.

For some units (e.g. frames) you will need to specify a sample rate:

Examples

julia> inseconds(50.0ms) 0.05

julia> inseconds(441frames, 44100Hz) 0.01

SampledSignals.kHz

SampledSignals.metadata — Function.

metadata(x, key::Symbol)
metadata(x, key::Symbol, idx)

Provide implementation-specific metadata for the given buffer or stream. For instance, data from a WAV file might have metadata that comes from extra chunks read from the file. If no idx is given then the first piece of metadata matching the key is returned. If there are multiple matches, the user can provide an index, or : to return a list of all matching metadata.

SampledSignals.mix — Function.

Mix the channels of the source array into the channels of the dest array, using coefficients from the mix matrix. To mix an M-channel buffer to a N-channel buffer, mix should be MxN. src and dest should not share memory.

SampledSignals.mix! — Function.

Mix the channels of the source array into the channels of the dest array, using coefficients from the mix matrix. To mix an M-channel buffer to a N-channel buffer, mix should be MxN. src and dest should not share memory.

SampledSignals.mono — Function.

Mix the channels of the src array into a mono array.

SampledSignals.mono! — Function.

Mix the channels of the src array into the mono dest array.

SampledSignals.ms

SampledSignals.nchannels

SampledSignals.nframes

SampledSignals.ns

SampledSignals.s

SampledSignals.samplerate

SampledSignals.samplerate!

SampledSignals.showchannels

SampledSignals.srname

SampledSignals.ticks

SampledSignals.toindex — Function.

toindex(buf::SampleBuf, I)

Convert the given index value to one that Base knows how to use natively for indexing

SampledSignals.typename

SampledSignals.unitname

SampledSignals.unsafe_read! — Function.

unsafe_read!(source::SampleSource, buf::Array, frameoffset, framecount)

Reads samples from the given source to the given array, assuming that the channel count, sampling rate, and element types are matching. This isn't called from user code, but is called by the read! (and likewise read) implementions in SampledSignals after it verifies that the buffer and sink are compatible, or possibly adds a conversion wrapper. SampledSignals will call this method with a 1D or 2D (nframes x nchannels) Array, with each channel in its own column. framecount frames of data should be copied into the array starting at frameoffset+1.

SampledSignals.unsafe_write — Function.

unsafe_write(sink::SampleSink, buf::Array, frameoffset, framecount)

Writes the given buffer to the given sink, assuming that the channel count, sampling rate, and element types are matching. This isn't called from user code, but is called by the write implemention in SampledSignals after it verifies that the buffer and sink are compatible, or possibly adds a conversion wrapper. sampledsignals will call this method with a 1D or 2D (nframes x nchannels) Array, with each channel in its own column. framecount frames of data should be copied from the array starting at frameoffset+1.

SampledSignals.wavwrite

SampledSignals.wrap_sink

SampledSignals.write_format

SampledSignals.write_header

SampledSignals.write_le

SampledSignals.μs