Docstrings

PyCall.@npyinitialize

PyCall.@py_str — Macro.

py".....python code....."

Evaluate the given Python code string in the main Python module.

If the string is a single line (no newlines), then the Python expression is evaluated and the result is returned. If the string is multiple lines (contains a newline), then the Python code is compiled and evaluated in the __main__ Python module and nothing is returned.

If the o option is appended to the string, as in py"..."o, then the return value is an unconverted PyObject; otherwise, it is automatically converted to a native Julia type if possible.

Any $var or $(expr) expressions that appear in the Python code (except in comments or string literals) are evaluated in Julia and passed to Python via auto-generated global variables. This allows you to "interpolate" Julia values into Python code.

Similarly, ny $$var or $$(expr) expressions in the Python code are evaluated in Julia, converted to strings via string, and are pasted into the Python code. This allows you to evaluate code where the code itself is generated by a Julia expression.

PyCall.@pycall — Macro.

@pycall func(args...)::T

Convenience macro which turns func(args...)::T into pycall(func, T, args...)

PyCall.@pycheck

PyCall.@pycheckn

PyCall.@pycheckv

PyCall.@pycheckz

PyCall.@pydef — Macro.

@pydef creates a Python class whose methods are implemented in Julia. For instance,

@pyimport numpy.polynomial as P
@pydef type Doubler <: P.Polynomial
    __init__(self, x=10) = (self[:x] = x)
    my_method(self, arg1::Number) = arg1 + 20
    x2.get(self) = self[:x] * 2
    x2.set!(self, new_val) = (self[:x] = new_val / 2)
end
Doubler()[:x2]

is essentially equivalent to the following Python code:

class Doubler(numpy.polynomial.Polynomial):
    def __init__(self, x=10):
        self.x = x
    def my_method(self, arg1): return arg1 + 20
    @property
    def x2(self): return self.x * 2
    @x2.setter
    def x2(self, new_val):
        self.x = new_val / 2
Doubler().x2

The method arguments and return values are automatically converted between Julia and Python. All Python special methods are supported (__len__, __add__, etc.)

@pydef allows for multiple inheritance of Python types:

@pydef type SomeType <: (BaseClass1, BaseClass2)
    ...
end

Multiple dispatch works, too:

x2.set!(self, new_x::Int) = ...
x2.set!(self, new_x::Float64) = ...

PyCall.@pydef_object — Macro.

@pydef_object is like @pydef, but it returns the metaclass as a PyObject instead of binding it to the class name. It's side-effect-free, except for the definition of the class methods.

PyCall.@pyglobal

PyCall.@pyglobalobj

PyCall.@pyglobalobjptr

PyCall.@pyimport

PyCall.@pylab

PyCall.@pysym

PyCall.@pywith — Macro.

@pywith

Mimics a Python 'with' statement. Usage:

@pywith EXPR[::TYPE1] [as VAR[::TYPE2]] begin BLOCK end

TYPE1/TYPE2 can optionally be used to override automatic conversion to Julia types for both the context manager and variable in cases where this is not desired.

PyCall.@return_not_None

PyCall.@with_ioraise

PyCall.ActivatePyActCtx

PyCall.BuiltinFunctionType

PyCall.DateTimeType

PyCall.DateTime_FromDateAndTime

PyCall.DateType

PyCall.Date_FromDate

PyCall.DeactivatePyActCtx

PyCall.DeltaType

PyCall.Delta_FromDelta

PyCall.Dl_info

PyCall.EnumProcessModules

PyCall.FuncWrapper

PyCall.LongInt

PyCall.METH_CLASS

PyCall.METH_KEYWORDS

PyCall.METH_NOARGS

PyCall.METH_O

PyCall.METH_STATIC

PyCall.METH_VARARGS

PyCall.MethodType

PyCall.MethodWrapperType

PyCall.NPY_ANYORDER

PyCall.NPY_ARRAY_ALIGNED

PyCall.NPY_ARRAY_C_CONTIGUOUS

PyCall.NPY_ARRAY_ELEMENTSTRIDES

PyCall.NPY_ARRAY_ENSUREARRAY

PyCall.NPY_ARRAY_ENSURECOPY

PyCall.NPY_ARRAY_FORCECAST

PyCall.NPY_ARRAY_F_CONTIGUOUS

PyCall.NPY_ARRAY_NOTSWAPPED

PyCall.NPY_ARRAY_OWNDATA

PyCall.NPY_ARRAY_UPDATEIFCOPY

PyCall.NPY_ARRAY_WRITEABLE

PyCall.NPY_BOOL

PyCall.NPY_BYTE

PyCall.NPY_CDOUBLE

PyCall.NPY_CFLOAT

PyCall.NPY_CLONGDOUBLE

PyCall.NPY_CORDER

PyCall.NPY_DOUBLE

PyCall.NPY_FLOAT

PyCall.NPY_FORTRANORDER

PyCall.NPY_HALF

PyCall.NPY_INT

PyCall.NPY_LONG

PyCall.NPY_LONGDOUBLE

PyCall.NPY_LONGLONG

PyCall.NPY_OBJECT

PyCall.NPY_SHORT

PyCall.NPY_STRING

PyCall.NPY_TYPES

PyCall.NPY_UBYTE

PyCall.NPY_UINT

PyCall.NPY_ULONG

PyCall.NPY_ULONGLONG

PyCall.NPY_UNICODE

PyCall.NPY_USHORT

PyCall.NPY_VOID

PyCall.NULL_UInt8_Ptr

PyCall.NpyArray

PyCall.PYTHONHOME

PyCall.PY_WRITE_RESTRICTED

PyCall.PyAny

PyCall.PyArray — Type.

PyArray(o::PyObject)

This converts an ndarray object o to a PyArray.

This implements a nocopy wrapper to a NumPy array (currently of only numeric types only).

If you are using pycall and the function returns an ndarray, you can use PyArray as the return type to directly receive a PyArray.

PyCall.PyArray_Info

PyCall.PyBUF_ANY_CONTIGUOUS

PyCall.PyBUF_C_CONTIGUOUS

PyCall.PyBUF_FORMAT

PyCall.PyBUF_F_CONTIGUOUS

PyCall.PyBUF_INDIRECT

PyCall.PyBUF_MAX_NDIM

PyCall.PyBUF_ND

PyCall.PyBUF_SIMPLE

PyCall.PyBUF_STRIDES

PyCall.PyBUF_WRITABLE

PyCall.PyBuffer

PyCall.PyCall

PyCall.PyDateTime_CAPI

PyCall.PyDateTime_Check

PyCall.PyDateTime_Delta

PyCall.PyDate_Check

PyCall.PyDate_HEAD

PyCall.PyDelta_Check

PyCall.PyDelta_FromDSU

PyCall.PyDict — Type.

PyDict(o::PyObject)
PyDict(d::Dict{K,V})

This returns a PyDict, which is a no-copy wrapper around a Python dictionary.

Alternatively, you can specify the return type of a pycall as PyDict.

PyCall.PyDict_Iterator

PyCall.PyError

PyCall.PyGetSetDef

PyCall.PyIO

PyCall.PyIOError

PyCall.PyMemberDef

PyCall.PyMethodDef

PyCall.PyNULL — Function.

PyNULL()

Return a PyObject that has a NULL underlying pointer, i.e. it doesn't actually refer to any Python object.

This is useful for initializing PyObject global variables and array elements before an actual Python object is available. For example, you might do const myglobal = PyNULL() and later on (e.g. in a module __init__ function), reassign myglobal to point to an actual object with copy!(myglobal, someobject). This procedure will properly handle Python's reference counting (so that the Python object will not be freed until you are done with myglobal).

PyCall.PyObject — Type.

PyObject(juliavar)

This converts a julia variable to a PyObject, which is a reference to a Python object. You can convert back to native julia types using convert(T, o::PyObject), or using PyAny(o).

Given o::PyObject, o[:attribute] is equivalent to o.attribute in Python, with automatic type conversion.

Given o::PyObject, get(o, key) is equivalent to o[key] in Python, with automatic type conversion.

PyCall.PyObject_struct

PyCall.PyPtr

PyCall.PyPtr_NULL

PyCall.PyReverseDims — Function.

PyReverseDims(array)

Passes a Julia array to Python as a NumPy row-major array (rather than Julia's native column-major order) with the dimensions reversed (e.g. a 2×3×4 Julia array is passed as a 4×3×2 NumPy row-major array). This is useful for Python libraries that expect row-major data.

PyCall.PyString_AsStringAndSize

PyCall.PyString_FromStringAndSize

PyCall.PyString_Size

PyCall.PyString_Type

PyCall.PyTextIO — Function.

PyTextIO(io::IO)
PyObject(io::IO)

Julia IO streams are converted into Python objects implementing the RawIOBase interface, so they can be used for binary I/O in Python

PyCall.PyTypeObject

PyCall.PyTypeObject!

PyCall.PyUnicode_AsUTF8String

PyCall.PyUnicode_DecodeUTF8

PyCall.PyVector — Type.

PyVector(o::PyObject)

This returns a PyVector object, which is a wrapper around an arbitrary Python list or sequence object.

Alternatively, PyVector can be used as the return type for a pycall that returns a sequence object (including tuples).

PyCall.Py_EQ

PyCall.Py_GE

PyCall.Py_GT

PyCall.Py_GetVersion

PyCall.Py_LE

PyCall.Py_LT

PyCall.Py_NE

PyCall.Py_SetPythonHome

PyCall.Py_TPFLAGS_BASETYPE

PyCall.Py_TPFLAGS_BASE_EXC_SUBCLASS

PyCall.Py_TPFLAGS_BYTES_SUBCLASS

PyCall.Py_TPFLAGS_CHECKTYPES

PyCall.Py_TPFLAGS_DICT_SUBCLASS

PyCall.Py_TPFLAGS_GC

PyCall.Py_TPFLAGS_HAVE_CLASS

PyCall.Py_TPFLAGS_HAVE_GC

PyCall.Py_TPFLAGS_HAVE_GETCHARBUFFER

PyCall.Py_TPFLAGS_HAVE_INDEX

PyCall.Py_TPFLAGS_HAVE_INPLACEOPS

PyCall.Py_TPFLAGS_HAVE_ITER

PyCall.Py_TPFLAGS_HAVE_NEWBUFFER

PyCall.Py_TPFLAGS_HAVE_RICHCOMPARE

PyCall.Py_TPFLAGS_HAVE_SEQUENCE_IN

PyCall.Py_TPFLAGS_HAVE_STACKLESS_EXTENSION

PyCall.Py_TPFLAGS_HAVE_STACKLESS_EXTENSION_

PyCall.Py_TPFLAGS_HAVE_VERSION_TAG

PyCall.Py_TPFLAGS_HAVE_WEAKREFS

PyCall.Py_TPFLAGS_HEAPTYPE

PyCall.Py_TPFLAGS_INT_SUBCLASS

PyCall.Py_TPFLAGS_IS_ABSTRACT

PyCall.Py_TPFLAGS_LIST_SUBCLASS

PyCall.Py_TPFLAGS_LONG_SUBCLASS

PyCall.Py_TPFLAGS_READY

PyCall.Py_TPFLAGS_READYING

PyCall.Py_TPFLAGS_STRING_SUBCLASS

PyCall.Py_TPFLAGS_TUPLE_SUBCLASS

PyCall.Py_TPFLAGS_TYPE_SUBCLASS

PyCall.Py_TPFLAGS_UNICODE_SUBCLASS

PyCall.Py_TPFLAGS_VALID_VERSION_TAG

PyCall.Py_buffer

PyCall.Py_eval_input

PyCall.Py_file_input

PyCall.Py_hash_t

PyCall.Py_jlWrap

PyCall.Py_single_input

PyCall.READONLY

PyCall.READ_RESTRICTED

PyCall.RESTRICTED

PyCall.T_BOOL

PyCall.T_BYTE

PyCall.T_CHAR

PyCall.T_DOUBLE

PyCall.T_FLOAT

PyCall.T_INT

PyCall.T_LONG

PyCall.T_LONGLONG

PyCall.T_NONE

PyCall.T_OBJECT

PyCall.T_OBJECT_EX

PyCall.T_PYSSIZET

PyCall.T_SHORT

PyCall.T_STRING

PyCall.T_STRING_INPLACE

PyCall.T_UBYTE

PyCall.T_UINT

PyCall.T_ULONG

PyCall.T_ULONGLONG

PyCall.T_USHORT

PyCall.Tk_eventloop

PyCall.TypeTuple

PyCall.__init__

PyCall.__pycall! — Function.

Lowest level version of pycall!(ret, o, ...), assumes pyargsptr and kw have all their args set to Python values, so we can just call the function o. Sets ret.o to the result of the call, and returns ret::PyObject.

PyCall._getindex

PyCall._localvar_counter

PyCall._maindict

PyCall._pickle

PyCall._ps_len

PyCall._ps_ptr

PyCall._pycall! — Function.

Low-level version of pycall!(ret, o, ...; kwargs...) Sets ret.o to the result of the call, and returns ret::PyObject

Low-level version of pycall!(ret, o, ...) for when kw is already in python friendly format but you don't have the python tuple to hold the arguments (pyargsptr). Sets ret.o to the result of the call, and returns ret::PyObject.

PyCall._pyimport

PyCall._pyjlwrap_call

PyCall._pywith

PyCall._start

PyCall.aligned

PyCall.anaconda_conda — Function.

anaconda_conda()

Return the path of the conda program if PyCall is configured to use an Anaconda install (but not the Conda.jl Python), and the empty string otherwise.

PyCall.array2py

PyCall.builtin

PyCall.c_contiguous

PyCall.c_void_p_Type

PyCall.callsym

PyCall.conda — Constant.

True if we are using the Python distribution in the Conda package.

PyCall.def_py_class — Function.

def_py_class(type_name::AbstractString, methods::Vector;
             base_classes=[], getsets::Vector=[])

def_py_class creates a Python class whose methods are implemented in Julia. @pydef macros expand into a call to this function.

Arguments

methods: a vector of tuples (py_name::String, jl_fun::Function) py_name will be a method of the Python class, which will call jl_fun

base_classes: the Python base classes to inherit from.

Return value: the created class (::PyTypeObject)

PyCall.delta_dsμ

PyCall.depfile

PyCall.dlinfo1

PyCall.dlinfo2

PyCall.docstring

PyCall.eval

PyCall.eventloops

PyCall.f_contiguous

PyCall.findsym

PyCall.fixqtpath

PyCall.format_traceback

PyCall.gstring_ptr

PyCall.gtk_eventloop

PyCall.gtk_requireversion

PyCall.gui

PyCall.hashsalt

PyCall.hassym

PyCall.include

PyCall.ind2py

PyCall.init_datetime

PyCall.inspect

PyCall.install_doevent

PyCall.interpolate_pycode

PyCall.ioraise

PyCall.is_mapping_object

PyCall.is_pyjlwrap

PyCall.iscontiguous

PyCall.ispybytearray

PyCall.ispynull — Function.

ispynull(o::PyObject)

Test where o contains a NULL pointer to a Python object, i.e. whether it is equivalent to a PyNULL() object.

PyCall.isseekable

PyCall.istuplen

PyCall.isvatuple

PyCall.jlWrapIteratorType

PyCall.jlWrapType

PyCall.jl_io_readline

PyCall.jl_io_readlines

PyCall.jl_io_seek

PyCall.jlfun2pyfun

PyCall.jlwrap_iterator

PyCall.julia_args

PyCall.julia_kwarg

PyCall.libpy_handle

PyCall.libpython

PyCall.maindict

PyCall.make_fname

PyCall.modulename

PyCall.mpc

PyCall.mpf

PyCall.mpmath

PyCall.mpmath_init

PyCall.mpprec

PyCall.npy_api

PyCall.npy_bool

PyCall.npy_complexfloating

PyCall.npy_floating

PyCall.npy_initialized

PyCall.npy_integer

PyCall.npy_multiarray

PyCall.npy_number

PyCall.npy_type

PyCall.npy_typestrs

PyCall.npyinitialize

PyCall.parse_pydef

PyCall.parse_pydef_toplevel

PyCall.permanent_strings

PyCall.pickle

PyCall.proc_handle

PyCall.py2array

PyCall.pyFSM

PyCall.pyany_toany

PyCall.pyarray_dims

PyCall.pybuiltin — Function.

pybuiltin(s::AbstractString)

Look up a string or symbol s among the global Python builtins. If s is a string it returns a PyObject, while if s is a symbol it returns the builtin converted to PyAny.

PyCall.pybytes — Function.

pybytes(b::Union{String,DenseVector{UInt8}})

Convert b to a Python bytes object. This differs from the default PyObject(b) conversion of String to a Python string (which may fail if b does not contain valid Unicode), or from the default conversion of a Vector{UInt8} to a bytearray object (which is mutable, unlike bytes).

PyCall.pycall — Function.

pycall(o::Union{PyObject,PyPtr}, returntype::TypeTuple, args...; kwargs...)

Call the given Python function (typically looked up from a module) with the given args... (of standard Julia types which are converted automatically to the corresponding Python types if possible), converting the return value to returntype (use a returntype of PyObject to return the unconverted Python object reference, or of PyAny to request an automated conversion)

PyCall.pycall! — Function.

pycall!(ret::PyObject, o::Union{PyObject,PyPtr}, returntype::Type, args...; kwargs...)

Set ret to the result of pycall(o, returntype, args...; kwargs) and return convert(returntype, ret). Avoids allocating an extra PyObject for ret. See pycall for other details.

PyCall.pycall_gc

PyCall.pycomplex_query

PyCall.pydate_query

PyCall.pydecref

PyCall.pydecref_

PyCall.pydict_query

PyCall.pyembed

PyCall.pyerr_check

PyCall.pyerr_clear

PyCall.pyerr_occurred

PyCall.pyeval — Function.

pyeval(s::AbstractString, returntype::TypeTuple=PyAny, locals=PyDict{AbstractString, PyObject}(),
                            input_type=Py_eval_input; kwargs...)

This evaluates s as a Python string and returns the result converted to rtype (which defaults to PyAny). The remaining arguments are keywords that define local variables to be used in the expression.

For example, pyeval("x + y", x=1, y=2) returns 3.

PyCall.pyeval_

PyCall.pyexc

PyCall.pyexc_initialize

PyCall.pyexists

PyCall.pyfloat_query

PyCall.pyfunction — Function.

pyfunction(f, argtypes...; kwtypes...)

Create a Python object that wraps around the Julia function (or callable object) f. Unlike PyObject(f), this allows you to specify the argument types that the Julia function expects — giving you more control and potentially better performance.

kwtypes... should be a set of somekeyword=SomeType arguments giving the desired Julia type for each keyword somekeyword. Unspecified keywords are converted to PyAny (i.e. auto-detected) by default.

The return value ret = f(...) is still converted back to a Python object by PyObject(ret). If you want a different return-type conversion than the default of PyObject(ret), you can instead call pyfunctionret.

PyCall.pyfunction_query

PyCall.pyfunctionret — Function.

pyfunctionret(f, returntype, argtypes...; kwtypes...)

Like pyfunction, but also lets you specify the returntype for conversion back to Python. In particular, if ret = f(...) is the return value of f, then it is converted to Python via PyObject(returntype(ret)).

If returntype is Any, then ret is not converted to a "native" Python type at all, and is instead converted to a "wrapped" Julia object in Python. If returntype is nothing, then the return value is discarded and nothing is returned to Python.

PyCall.pygui — Function.

pygui()

Return the current GUI toolkit as a symbol.

PyCall.pygui_start — Function.

pygui_start(gui::Symbol = pygui())

Start the event loop of a certain toolkit.

The argument gui defaults to the current default GUI, but it could be :wx, :gtk, :gtk3, :tk, or :qt.

PyCall.pygui_stop — Function.

pygui_stop(gui::Symbol = pygui())

Stop any running event loop for gui. The gui argument defaults to current default GUI.

PyCall.pygui_stop_all

PyCall.pygui_works

PyCall.pyimport — Function.

pyimport(s::AbstractString)

Import the Python module s (a string or symbol) and return a pointer to it (a PyObject). Functions or other symbols in the module may then be looked up by s[name] where name is a string (for the raw PyObject) or symbol (for automatic type-conversion). Unlike the @pyimport macro, this does not define a Julia module and members cannot be accessed with s.name

PyCall.pyimport_conda — Function.

pyimport_conda(modulename, condapkg, [channel])

Returns the result of pyimport(modulename) if possible. If the module is not found, and PyCall is configured to use the Conda Python distro (via the Julia Conda package), then automatically install condapkg via Conda.add(condapkg) and then re-try the pyimport. Other Anaconda-based Python installations are also supported as long as their conda program is functioning.

If PyCall is not using Conda and the pyimport fails, throws an exception with an error message telling the user how to configure PyCall to use Conda for automated installation of the module.

The third argument, channel is an optional Anaconda "channel" to use for installing the package; this is useful for packages that are not included in the default Anaconda package listing.

PyCall.pyimport_e — Function.

pyimport_e(s::AbstractString)

Like pyimport(s), but returns an ispynull(o) == true object if the module cannot be imported rather than throwing an error.

PyCall.pyimport_name

PyCall.pyincref

PyCall.pyincref_

PyCall.pyint_query

PyCall.pyio_initialize

PyCall.pyio_initialized

PyCall.pyio_jl

PyCall.pyisinstance

PyCall.pyjlwrap_call

PyCall.pyjlwrap_dealloc

PyCall.pyjlwrap_doc

PyCall.pyjlwrap_getattr

PyCall.pyjlwrap_getiter

PyCall.pyjlwrap_hash

PyCall.pyjlwrap_hash32

PyCall.pyjlwrap_init

PyCall.pyjlwrap_iternext

PyCall.pyjlwrap_membername

PyCall.pyjlwrap_members

PyCall.pyjlwrap_new

PyCall.pyjlwrap_repr

PyCall.pyjlwrap_type

PyCall.pyjlwrap_type!

PyCall.pymp_query

PyCall.pynothing

PyCall.pynothing_query

PyCall.pyprogramname

PyCall.pyproperty

PyCall.pyptr_query

PyCall.pyquery

PyCall.pyraise

PyCall.pyrepr — Function.

pyrepr(o::PyObject)

Return a string representation of o corresponding to repr(o) in Python.

PyCall.pyreturn — Function.

pyreturn(x) :: PyPtr

Prepare PyPtr from x for passing it to Python. If x is already a PyObject, the refcount is incremented. Otherwise a PyObject wrapping/converted from x is created.

PyCall.pysalt

PyCall.pysalt32

PyCall.pysequence_query

PyCall.pysetarg!

PyCall.pysetargs!

PyCall.pystealref! — Function.

"Steal" a reference from a PyObject: return the raw PyPtr, while setting the corresponding o.o field to NULL so that no decref will be performed when o is garbage collected. (This means that you can no longer use o.) Used for passing objects to Python.

PyCall.pystr — Function.

pystr(o::PyObject)

Return a string representation of o corresponding to str(o) in Python.

PyCall.pystring — Function.

pystring(o::PyObject)

Return a string representation of o. Normally, this corresponds to repr(o) in Python, but unlike repr it should never fail (falling back to str or to printing the raw pointer if necessary).

PyCall.pystring_query

PyCall.python

PyCall.pytype_mapping — Function.

pytype_mapping(pytype, jltype)

Given a Python type object pytype, tell PyCall to convert it to jltype in PyAny(object) conversions.

PyCall.pytype_queries

PyCall.pytype_query — Function.

pytype_query(o::PyObject, default=PyObject)

Given a Python object o, return the corresponding native Julia type (defaulting to default) that we convert o to in PyAny(o) conversions.

PyCall.pytypeof

PyCall.pyunicode_literals

PyCall.pyversion

PyCall.pyversion_build

PyCall.pywrap — Function.

pywrap(o::PyObject)

This returns a wrapper w that is an anonymous module which provides (read) access to converted versions of o's members as w.member.

For example, @pyimport module as name is equivalent to const name = pywrap(pyimport("module"))

If the Python module contains identifiers that are reserved words in Julia (e.g. function), they cannot be accessed as w.member; one must instead use w.pymember(:member) (for the PyAny conversion) or w.pymember("member") (for the raw PyObject).

PyCall.pywrapfn

PyCall.pyxrange

PyCall.qt_eventloop

PyCall.reserved

PyCall.set!

PyCall.sizeof_PyObject_HEAD

PyCall.some_address_in_libpython

PyCall.some_address_in_main_exe

PyCall.symbols_present

PyCall.tkinter_name

PyCall.tuptype

PyCall.typetuple

PyCall.ufuncType

PyCall.unsafe_pyjlwrap_to_objref

PyCall.wPYTHONHOME

PyCall.weakref_callback

PyCall.weakref_callback_meth

PyCall.weakref_callback_obj

PyCall.wpyprogramname

PyCall.writedims

PyCall.writeok_assign

PyCall.wx_eventloop

PyCall.xrange