Docstrings

IJulia.@verror_show

IJulia.@vprintln

IJulia.CommManager

IJulia.IJULIA_DEBUG

IJulia.IJulia — Module.

IJulia is a Julia-language backend combined with the Jupyter interactive environment (also used by IPython). This combination allows you to interact with the Julia language using Jupyter/IPython's powerful graphical notebook, which combines code, formatted text, math, and multimedia in a single document.

The IJulia module is used in three ways

Typing using IJulia; notebook() will launch the Jupyter notebook interface in your web browser. This is an alternative to launching jupyter notebook directly from your operating-system command line.

In a running notebook, the IJulia module is loaded and IJulia.somefunctions can be used to interact with the running IJulia kernel:

IJulia.load(filename) and IJulia.load_string(s) load the contents of a file or a string, respectively, into a notebook cell.

IJulia.clear_output() to clear the output from the notebook cell, useful for simple animations.

IJulia.clear_history() to clear the history variables In and Out.

push_X_hook(f) and pop_X_hook(f), where X is either preexecute, postexecute, or posterror. This allows you to insert a "hook" function into a list of functions to execute when notebook cells are evaluated.

IJulia.set_verbose() enables verbose output about what IJulia is doing internally; this is mainly used for debugging.

It is used internally by the IJulia kernel when talking to the Jupyter server.

IJulia.IJuliaStdio

IJulia.IJulia_RNG

IJulia.In — Constant.

In is a global dictionary of input strings, where In[n] returns the string for input cell n of the notebook (as it was when it was last evaluated).

IJulia.InlineDisplay

IJulia.JUPYTER

IJulia.MiniREPL

IJulia.Msg

IJulia.Out — Constant.

Out is a global dictionary of output values, where Out[n] returns the output from the last evaluation of cell n in the notebook.

IJulia.SOFTSCOPE

IJulia.__init__

IJulia._showable

IJulia.alias_magic_help

IJulia.ans — Constant.

ans is a global variable giving the value returned by the last notebook cell evaluated.

IJulia.bufs

IJulia.capture_stderr

IJulia.capture_stdout — Constant.

The IJulia kernel captures all stdout and stderr output and redirects it to the notebook. When debugging IJulia problems, however, it can be more convenient to not capture stdout and stderr output (since the notebook may not be functioning). This can be done by editing IJulia.jl to set capture_stderr and/or capture_stdout to false.

IJulia.cd_magic_help

IJulia.check_prompt_streams

IJulia.chr2ind

IJulia.clear_history — Function.

clear_history([indices])

The clear_history() function clears all of the input and output history stored in the running IJulia notebook. This is sometimes useful because all cell outputs are remember in the Out global variable, which prevents them from being freed, so potentially this could waste a lot of memory in a notebook with many large outputs.

The optional indices argument is a collection of indices indicating a subset of cell inputs/outputs to clear.

IJulia.clear_output — Function.

clear_output(wait=false)

Call clear_output() to clear visible output from the current notebook cell. Using wait=true clears the output only when new output is available, which reduces flickering and is useful for simple animations.

IJulia.complete_request

IJulia.complete_type

IJulia.complete_types

IJulia.connect_request

IJulia.connection_file

IJulia.control

IJulia.copy_config

IJulia.current_module

IJulia.debug_magic_help

IJulia.default_jupyter_data_dir

IJulia.depfile

IJulia.display_dict — Function.

Generate a dictionary of mime_type => data pairs for all registered MIME types. This is the format that Jupyter expects in displaydata and executeresult messages.

IJulia.display_mimejson — Function.

Generate the preferred json-MIME representation of x.

Returns a tuple with the selected MIME type and the representation of the data using that MIME type (as a JSONText).

IJulia.display_mimestring — Function.

Generate the preferred MIME representation of x.

Returns a tuple with the selected MIME type and the representation of the data using that MIME type.

IJulia.displayable

IJulia.displayqueue

IJulia.docdict

IJulia.edit_magic_help

IJulia.env_magic_help

IJulia.error_content

IJulia.eval

IJulia.eventloop

IJulia.exe

IJulia.execute_msg

IJulia.execute_payloads

IJulia.execute_request

IJulia.file_magic_help

IJulia.find_jupyter_subcommand

IJulia.find_parsestart

IJulia.flush_all

IJulia.generic_magic_help

IJulia.get_log_preface

IJulia.get_token

IJulia.gui_magic_help

IJulia.handlers

IJulia.heartbeat

IJulia.heartbeat_thread

IJulia.history — Function.

history([io], [indices...])

The history() function prints all of the input history stored in the running IJulia notebook in a format convenient for copying.

The optional indices argument is one or more indices or collections of indices indicating a subset input cells to print.

The optional io argument is for specifying an output stream. The default is stdout.

IJulia.history_magic_help

IJulia.history_request

IJulia.hmac

IJulia.hmacstate

IJulia.html_magic_help

IJulia.ijulia_jsonmime_types — Constant.

MIME types that when rendered (via stringmime) return JSON data. See ijulia_mime_types for a description of how MIME types are selected.

This is necessary to embed the JSON as is in the displaydata bundle (rather than as stringify'd JSON).

IJulia.ijulia_mime_types — Constant.

A vector of MIME types (or vectors of MIME types) that IJulia will try to render. IJulia will try to render every MIME type specified in the first level of the vector. If a vector of MIME types is specified, IJulia will include only the first MIME type that is renderable (this allows for the expression of priority and exclusion of redundant data).

For example, since "text/plain" is specified as a first-child of the array, IJulia will always try to include a "text/plain" representation of anything that is displayed. Since markdown and html are specified within a sub-vector, IJulia will always try to render "text/markdown", and will only try to render "text/html" if markdown isn't possible.

IJulia.include

IJulia.ind2chr

IJulia.ind_to_utf16

IJulia.init

IJulia.inited — Constant.

inited is a global variable that is set to true if the IJulia kernel is running, i.e. in a running IJulia notebook. To test whether you are in an IJulia notebook, therefore, you can check isdefined(Main, :IJulia) && IJulia.inited.

IJulia.inspect_request

IJulia.installkernel — Function.

installkernel(name::AbstractString, options::AbstractString...;
              specname::AbstractString,
              env=Dict())

Install a new Julia kernel, where the given options are passed to the julia executable, the user-visible kernel name is given by name followed by the Julia version, and the env dictionary is added to the environment.

The new kernel name is returned by installkernel. For example:

kernelpath = installkernel("Julia O3", "-O3", env=Dict("FOO"=>"yes"))

creates a new Julia kernel in which julia is launched with the -O3 optimization flag and FOO=yes is included in the environment variables.

The returned kernelpath is the path of the installed kernel directory, something like /...somepath.../kernels/julia-O3-1.0 (in Julia 1.0). The specname argument can be passed to alter the name of this directory (which defaults to name with spaces replaced by hyphens).

You can uninstall the kernel by calling rm(kernelpath, recursive=true).

IJulia.interrupt_request

IJulia.ipy_mime

IJulia.is_complete_request

IJulia.israwtext

IJulia.isyes

IJulia.javascript_magic_help

IJulia.jupyter_data_dir

IJulia.jupyterlab — Function.

jupyterlab(; dir=homedir(), detached=false)

Similar to IJulia.notebook() but launches JupyterLab instead of the Jupyter notebook.

IJulia.kernel_info_request

IJulia.kerneldir

IJulia.latex_magic_help

IJulia.launch

IJulia.limitstringmime

IJulia.load — Function.

load(filename, replace=false)

Load the file given by filename into a new input code cell in the running IJulia notebook, analogous to the %load magics in IPython. If the optional argument replace is true, then the file contents replace the current cell rather than creating a new cell.

IJulia.load_magic_help

IJulia.load_string — Function.

load_string(s, replace=false)

Load the string s into a new input code cell in the running IJulia notebook, somewhat analogous to the %load magics in IPython. If the optional argument replace is true, then s replaces the current cell rather than creating a new cell.

IJulia.lsmagic_help

IJulia.magic_help

IJulia.magic_help_string

IJulia.magics_help

IJulia.magics_regex

IJulia.matplotlib_magic_help

IJulia.max_bytes

IJulia.max_output_per_request

IJulia.metadata

IJulia.minirepl

IJulia.msg_header

IJulia.msg_pub

IJulia.msg_reply

IJulia.n — Constant.

IJulia.n is the (integer) index of the last-evaluated notebook cell.

IJulia.notebook — Function.

notebook(; dir=homedir(), detached=false)

The notebook() function launches the Jupyter notebook, and is equivalent to running jupyter notebook at the operating-system command-line. The advantage of launching the notebook from Julia is that, depending on how Jupyter was installed, the user may not know where to find the jupyter executable.

By default, the notebook server is launched in the user's home directory, but this location can be changed by passing the desired path in the dir keyword argument. e.g. notebook(dir=pwd()) to use the current directory.

By default, notebook() does not return; you must hit ctrl-c or quit Julia to interrupt it, which halts Jupyter. So, you must leave the Julia terminal open for as long as you want to run Jupyter. Alternatively, if you run notebook(detached=true), the jupyter notebook will launch in the background, and will continue running even after you quit Julia. (The only way to stop Jupyter will then be to kill it in your operating system's process manager.)

IJulia.num_utf8_trailing — Function.

If d ends with an incomplete UTF8-encoded character, return the number of trailing incomplete bytes. Otherwise, return 0.

IJulia.orig_stderr

IJulia.orig_stdin

IJulia.orig_stdout

IJulia.oslibuv_flush

IJulia.parseok

IJulia.paste_magic_help

IJulia.pdef_magic_help

IJulia.pdoc_magic_help

IJulia.pfile_magic_help

IJulia.pipe_magic_help

IJulia.pop_posterror_hook — Function.

pop_posterror_hook(f::Function)

Remove a function f() from the list of functions to execute after an error occurs when a notebook cell is evaluated.

IJulia.pop_postexecute_hook — Function.

pop_postexecute_hook(f::Function)

Remove a function f() from the list of functions to execute after executing any notebook cell.

IJulia.pop_preexecute_hook — Function.

pop_preexecute_hook(f::Function)

Remove a function f() from the list of functions to execute before executing any notebook cell.

IJulia.posterror_hooks

IJulia.postexecute_hooks

IJulia.precision_magic_help

IJulia.preexecute_hooks

IJulia.profile

IJulia.prun_magic_help

IJulia.psearch_magic_help

IJulia.publish

IJulia.push_posterror_hook — Function.

pop_posterror_hook(f::Function)

Remove a function f() from the list of functions to execute after an error occurs when a notebook cell is evaluated.

IJulia.push_postexecute_hook — Function.

push_postexecute_hook(f::Function)

Push a function f() onto the end of a list of functions to execute after executing any notebook cell.

IJulia.push_preexecute_hook — Function.

push_preexecute_hook(f::Function)

Push a function f() onto the end of a list of functions to execute before executing any notebook cell.

IJulia.pwd_magic_help

IJulia.qtconsole

IJulia.qtconsole_magic_help

IJulia.raw_input

IJulia.read_stderr

IJulia.read_stdout

IJulia.readprompt — Function.

readprompt(prompt::AbstractString; password::Bool=false)

Display the prompt string, request user input, and return the string entered by the user. If password is true, the user's input is not displayed during typing.

IJulia.recall_magic_help

IJulia.recv_ipython

IJulia.register_jsonmime

IJulia.register_mime

IJulia.requests

IJulia.requests_task

IJulia.run_magic_help

IJulia.save_magic_help

IJulia.sc_magic_help

IJulia.send_ipython

IJulia.send_status

IJulia.send_stderr

IJulia.send_stdio

IJulia.send_stdout

IJulia.send_stream

IJulia.set_cur_msg — Function.

Jupyter associates cells with message headers. Once a cell's execution state has been set as to idle, it will silently drop stream messages (i.e. output to stdout and stderr) - see https://github.com/jupyter/notebook/issues/518. When using Interact, and a widget's state changes, a new message header is sent to the IJulia kernel, and while Reactive is updating Signal graph state, it's execution state is busy, meaning Jupyter will not drop stream messages if Interact can set the header message under which the stream messages will be sent. Hence the need for this function.

IJulia.set_current_module

IJulia.set_env_magic_help

IJulia.set_max_stdio — Function.

set_max_stdio(max_output::Integer)

Sets the maximum number of bytes, max_output, that can be written to stdout and stderr before getting truncated. A large value here allows a lot of output to be displayed in the notebook, potentially bogging down the browser.

IJulia.set_verbose — Function.

set_verbose(v=true)

This function enables (or disables, for set_verbose(false)) verbose output from the IJulia kernel, when called within a running notebook. This consists of log messages printed to the terminal window where jupyter was launched, displaying information about every message sent or received by the kernel. Used for debugging IJulia.

IJulia.show_bt

IJulia.showerror_nobt

IJulia.shutdown_request

IJulia.socket_locks

IJulia.start_heartbeat

IJulia.stdio_bytes

IJulia.stream_interval

IJulia.svg_magic_help

IJulia.sx_magic_help

IJulia.threadid

IJulia.time_magic_help

IJulia.undisplay

IJulia.utf16_to_ind

IJulia.uuid4

IJulia.verbose

IJulia.waitloop

IJulia.watch_stdio

IJulia.watch_stream — Function.

Continually read from (size limited) Libuv/OS buffer into an IObuffer to avoid problems when the Libuv/OS buffer gets full (https://github.com/JuliaLang/julia/issues/8789). Send data immediately when buffer contains more than max_bytes bytes. Otherwise, if data is available it will be sent every stream_interval seconds (see the Timers set up in watchstdio). Truncate the output to `maxoutputperrequest` bytes per execution request since excessive output can bring browsers to a grinding halt.

IJulia.who_magic_help

IJulia.writefile_magic_help