Docstrings

WebSockets.@wslog

WebSockets.CLOSED

WebSockets.CLOSING

WebSockets.CONNECTED

WebSockets.Dt

WebSockets.OPCODE_BINARY — Constant.

%x2 denotes a binary frame

WebSockets.OPCODE_CLOSE — Constant.

%x8 denotes a connection close

WebSockets.OPCODE_CONTINUATION — Constant.

%x0 denotes a continuation frame

WebSockets.OPCODE_PING — Constant.

%x9 denotes a ping

WebSockets.OPCODE_PONG — Constant.

%xA denotes a pong

WebSockets.OPCODE_TEXT — Constant.

%x1 denotes a text frame

WebSockets.ReadyState

WebSockets.SUBProtocols — Constant.

Handshakes with subprotocols are rejected by default. Add to acceptable SUBProtocols through e.g.

   addsubproto("json")

Also see function subprotocol

WebSockets.ServerOptions

WebSockets.ServerWS — Type.

WebSockets.ServerWS(handler::Function, wshandler::Function, logger::IO)

WebSockets.ServerWS is an argument type for WebSockets.serve. Instances include .in and .out channels, see WebSockets.serve.

Server options can be set using keyword arguments, see methods(WebSockets.ServerWS)

WebSockets.TIMEOUT_CLOSEHANDSHAKE — Constant.

A reasonable amount of time

WebSockets.WebSocket — Type.

A WebSocket is a wrapper over a TCPSocket. It takes care of wrapping outgoing data in a frame and unwrapping (and concatenating) incoming data.

WebSockets.WebSocketClosedError

WebSockets.WebSocketError

WebSockets.WebSocketFragment — Type.

Represents one (received) message frame.

WebSockets.WebSocketLogger — Type.

Differences to stdlib/Logging/ConsoleLogger:

- default timestamp on logging messages (except @info)
- a 'shouldlog' function can be passed in. The `shouldlog_default` function filters
    on HTTP.Servers messages as well as on message_limits
- :wslog => true flag which may be used for context-sensitive output
    from 'show' methods. This means a user can define 'show' methods
    which are used with this logger without affecting the behaviour
    defined in other modules.
- :limited => true is included in the default IOContext. Keyword: show_limited
- string_with_env_ws is exported for easy overloading on specific types
- @info, @debug, @warn etc. will splat the first argument if it's a tuple arguments, e.g.

julia> var = "a"
"a"
julia> @info (1, var)
[ Info: 1a

WebSockets.WebSockets — Module.

WebSockets

This module implements the WebSockets protocol. It relies on the package HTTP.jl.

Websocket|server relies on a client initiating the connection. Websocket|client initiate the connection.

The client side of the connection is most typically a browser with scripts enabled. Browsers are always the initiating, client side. But the peer can be any program, in any language, that follows the protocol. That includes another Julia session, running in a parallel process or task.

Future improvements:

Check rsv1 to rsv3 values. This will reduce bandwidth.

Optimize maskswitch!, possibly threaded above a certain limit.

Split messages over several frames.

WebSockets.WebsocketHandler — Type.

WebsocketHandler(f::Function) <: Handler

The provided argument should be one of the forms f(WebSocket) => nothing f(Request, WebSocket) => nothing The latter form is intended for gatekeeping, ref. RFC 6455 section 10.1

f accepts a WebSocket and does interesting things with it, like reading, writing and exiting when finished.

WebSockets.Wslog

WebSockets._openstream — Function.

Called by open with a stream connected to a server, after handshake is initiated

WebSockets._show

WebSockets._uv_status_tuple — Function.

For colorful printing

WebSockets.addsubproto — Function.

Used to specify acceptable subprotocols. See SUBProtocols

WebSockets.check_upgrade — Function.

Throws WebSocketError if the upgrade message is not basically valid. Called from 'upgrade' for potential server side websockets, and from `_openstream' for potential client side websockets. Not normally called from user code.

WebSockets.checkratelimit! — Function.

'checkratelimit!' updates a dictionary of IP addresses which keeps track of their connection quota per time window.

The allowed connections per time is given in keyword argument ratelimit.

The actual ratelimit::Rational value, is normally given as a field value in ServerOpions.

'checkratelimit!' is the default rate limiting function for ServerWS, which passes it as the 'tcpisvalid' argument to 'WebSockets.HTTP.listen'. Other functions can be given as a keyword argument, as long as they adhere to this form, which WebSockets.HTTP.listen expects.

Base.close — Function.

close(ws::WebSocket)
close(ws::WebSocket, statusnumber = n)
close(ws::WebSocket, statusnumber = n, freereason = "my reason")

Send an OPCODE_CLOSE frame, and wait for the same response or until a reasonable amount of time, 10.0 s, has passed. Data received while closing is dropped. Status number n according to RFC 6455 7.4.1 can be included, see WebSockets.codeDesc

WebSockets.codeDesc — Constant.

Status codes according to RFC 6455 7.4.1

WebSockets.default_metaformat — Function.

Defines a default logging message format with timestamp

WebSockets.eval

WebSockets.generate_websocket_key — Function.

WebSocket Handshake Procedure

generate_websocket_key(key) transforms a websocket client key into the server's accept value. This is done in three steps:

Concatenate key with magic string from RFC.

SHA1 hash the resulting base64 string.

Encode the resulting number in base64.

This function then returns the string of the base64-encoded value.

WebSockets.handle_control_frame — Function.

Respond to pings, ignore pongs, respond to close.

WebSockets.hasprotocol — Function.

Used in handshake. See SUBProtocols

WebSockets.include

WebSockets.init_socket — Function.

Buffer writes to socket till flush (sock)

WebSockets.is_control_frame — Function.

Control frames have opcodes with the highest bit = 1.

WebSockets.is_upgrade — Function.

Fast checking for websocket upgrade request vs content requests. Called on all new connections in '_servercoroutine'.

WebSockets.locked_write — Function.

Write without interruptions

WebSockets.maskswitch! — Function.

maskswitch!(data)
maskswitch!(data, key:: 4-element Vector{UInt8})

Masks or unmasks data in-place, returns the key used. Calling twice with the same key restores data. Ref. RFC 6455 5-3.

WebSockets.open — Function.

Initiate a websocket|client connection to server defined by url. If the server accepts the connection and the upgrade to websocket, f is called with an open websocket|client

e.g. say hello, close and leave

using WebSockets
WebSockets.open("ws://127.0.0.1:8000") do ws
    write(ws, "Hello")
    println("that's it")
end;

If a server is listening and accepts, "Hello" is sent (as a Vector{UInt8}).

On exit, a closing handshake is started. If the server is not currently reading (which is a blocking function), this side will reset the underlying connection (ECONNRESET) after a reasonable amount of time and continue execution.

WebSockets.origin — Function.

origin(request) => String Convenience function for checking which server / port address the client claims its code was downloaded from. The resource path can be found with target(req). E.g.

function gatekeeper(req, ws)
    orig = WebSockets.origin(req)
        if startswith(orig, "http://localhost") || startswith(orig, "http://127.0.0.1")
            handlewebsocket(ws)
        end
    end
end

Base.read — Function.

read(ws::WebSocket)

Typical use: msg = String(read(ws)) Read one non-control message from a WebSocket. Any control messages that are read will be handled by the handlecontrolframe function. Only the data (contents/body/payload) of the message will be returned as a Vector{UInt8}.

This function will not return until a full non-control message has been read.

WebSockets.read_frame — Function.

Read a frame: turn bytes from the websocket into a WebSocketFragment.

WebSockets.readframe_nonblocking — Function.

For the closing handshake, we won't wait indefinitely for non-responsive clients. Returns a throwaway frame if the socket happens to be empty

WebSockets.readguarded — Function.

readguarded(websocket) => (Vector, Bool)

Return (data::Vector, true) or (Vector{UInt8}(), false)

The peer can potentially disconnect at any time, but no matter the cause you will usually just want to exit your websocket handling function when you can't write to it.

E.g.

while true
    data, success = readguarded(websocket)
    !success && break
    println(String(data))
end

To check the errors (if you get any), temporarily set loging min_level to Logging.debug, e.g:

using WebSockets, Logging
global_logger(WebSocketLogger(stderr, Logging.Debug));

WebSockets.send_ping — Function.

Send a ping message, optionally with data.

WebSockets.send_pong — Function.

Send a pong message, optionally with data.

WebSockets.serve — Function.

WebSockets.serve(server::ServerWS, port)
WebSockets.serve(server::ServerWS, host, port)
WebSockets.serve(server::ServerWS, host, port, verbose)

A wrapper for WebSockets.HTTP.listen. Puts any caught error and stacktrace on the server.out channel. To stop a running server, put a byte on the server.in channel.

    @async WebSockets.serve(server, "127.0.0.1", 8080)

After a suspected connection task failure:

    if isready(myserver_WS.out)
        stack_trace = take!(myserver_WS.out)
    end

WebSockets.shouldlog_default — Function.

Early filtering of messages based on message id limits, silencing of HTTP.Servers if defined

WebSockets.string_with_env_ws — Function.

Context-aware text representation of the first argument to logging macros. Made easily available for overloading on specific types.

WebSockets.subprotocol — Function.

subprotocol(request) => String

Convenience function for reading upgrade request subprotocol. Acceptable subprotocols need to be predefined using addsubproto(myprotocol). No other subprotocols will pass the handshake. E.g.

WebSockets.addsubproto("instructions")
WebSockets.addsubproto("relay_backend")
function gatekeeper(req, ws)
    subpr = WebSockets.subprotocol(req)
    if subpr == "instructions"
        instructions(ws)
    elseif subpr == "relay_backend"
        relay_backend(ws)
    end
end

Then, in browser javascript (or equivalent with Julia WebSockets.open( , ))

function load(){
    var wsuri = document.URL.replace("http:", "ws:");
    ws1 = new WebSocket(wsuri, "instructions");
    ws2 = new WebSocket(wsuri, "relay_backend");
    ws1.onmessage = function(e){doinstructions(e.data)};
    ...
    } // load

WebSockets.target — Function.

target(request) => String

Convenience function for reading upgrade request target. E.g.

    function gatekeeper(req, ws)
        if target(req) == "/gamepad"
            @spawnat 2 gamepad(ws)
        elseif target(req) == "/console"
            @spawnat 3 console(ws)
            ...
        end
    end

Then, in browser javascript (or equivalent with Julia WebSockets.open( , ))

function load(){
    var wsuri = document.URL.replace("http:", "ws:");
    ws1 = new WebSocket(wsuri + "/gamepad");
    ws2 = new WebSocket(wsuri + "/console");
    ws3 = new WebSocket(wsuri + "/graphics");
    ws4 = new WebSocket(wsuri + "/audiochat");
    ws1.onmessage = function(e){vibrate(e.data)}
    } // load

WebSockets.upgrade — Function.

Used as part of a server definition. Call this if is_upgrade(stream.message) returns true.

Responds to a WebSocket handshake request. If the connection is acceptable, sends status code 101 and headers according to RFC 6455, then calls user's handler function f with the connection wrapped in a WebSocket instance.

f(ws) is called with the websocket and no client info f(headers, ws) also receives a dictionary of request headers for added security measures

On exit from f, a closing handshake is started. If the client is not currently reading (which is a blocking function), this side will reset the underlying connection (ECONNRESET) after a reasonable amount of time and continue execution.

If the upgrade is not accepted, responds to client with '400'.

e.g. server with local error handling. Combine with WebSocket.open example.

using WebSockets

badgatekeeper(reqdict, ws) = sqrt(-2)
handlerequest(req) = WebSockets.Response(501)
const SERVERREF = Ref{Base.IOServer}()
try
    WebSockets.HTTP.listen("127.0.0.1", UInt16(8000), tcpref = SERVERREF) do stream
        if WebSockets.is_upgrade(stream.message)
            WebSockets.upgrade(badgatekeeper, stream)
        else
            WebSockets.handle_request(handlerequest, stream)
        end
    end
catch err
    showerror(stderr, err)
    println.(stacktrace(catch_backtrace())[1:4])
end

Base.write — Function.

Write text data; will be sent as one frame.

Write binary data; will be sent as one frame.

WebSockets.write_fragment — Function.

write_fragment(io, islast, opcode, hasmask, data::Array{UInt8})

Write the raw frame to a bufffer. Websocket|client must set 'hasmask'.

WebSockets.write_ping

WebSockets.write_pong

WebSockets.writeguarded — Function.

writeguarded(websocket, message) => Bool

Return true if write is successful, false if not. The peer can potentially disconnect at any time, but no matter the cause you will usually just want to exit your websocket handling function when you can't write to it.

To check the errors (if you get any), temporarily set loging min_level to Logging.debug, e.g:

using WebSockets, Logging
global_logger(WebSocketLogger(stderr, Logging.Debug));