Docstrings

Conda.BINDIR

Conda.CONDARC

Conda.Conda — Module.

The Conda module provides access to the conda packages manager to install binary dependencies of other Julia packages.

The main functions in Conda are:

Conda.add(package): install a package;

Conda.rm(package): remove (uninstall) a package;

Conda.update(): update all installed packages to the latest version;

Conda.list(): list all installed packages.

Conda.add_channel(channel): add a channel to the list of channels;

Conda.channels(): get the current list of channels;

Conda.rm_channel(channel): remove a channel from the list of channels;

```

Conda.Environment

Conda.LIBDIR

Conda.MINICONDA_VERSION

Conda.PREFIX

Conda.PYTHONDIR

Conda.PkgOrPkgs

Conda.ROOTENV

Conda.SCRIPTDIR

Conda._install_conda — Function.

Install miniconda if it hasn't been installed yet; installconda(true) installs Conda even if it has already been installed.

Conda._installed_packages — Function.

List all installed packages as an array.

Conda._installed_packages_dict — Function.

List all installed packages as an dict of tuples with (version_number, fullname).

Conda._installer_url — Function.

Get the miniconda installer URL.

Conda._quiet — Function.

Suppress progress bar in continuous integration environments

Conda._set_conda_env — Function.

Use a cleaned up environment for the command cmd.

Any environment variable starting by CONDA or PYTHON will interact with the run.

Conda.add — Function.

Install a new package or packages.

Conda.add_channel — Function.

Add a channel to the list of channels

Conda.bin_dir — Function.

Prefix for the executable files installed with the packages

Conda.channels — Function.

Get the list of channels used to search packages

Conda.clean — Function.

clean(;
    debug=false, index=true, locks=true, tarballs=true, packages=true, sources=true
)

Runs conda clean -y with the specified flags.

Conda.conda

Conda.conda_rc — Function.

Path to the condarc file

Conda.deps_file

Conda.eval

Conda.exists — Function.

Check if a given package exists.

Conda.export_list — Function.

export_list(filepath, env=/home/pfitzseb/.julia/conda/3)
export_list(io, env=/home/pfitzseb/.julia/conda/3)

List all packages and write them to an export file for use the Conda.import_list

Conda.import_list — Function.

" importlist(filename, env=/home/pfitzseb/.julia/conda/3, channels=String[]) importlist(io, env=/home/pfitzseb/.julia/conda/3, channels=String[])

Create a new environment with various channels and a packages list file.

Conda.include

Conda.lib_dir — Function.

Prefix for the shared libraries installed with the packages

Conda.list — Function.

List all installed packages to standard output.

Conda.parseconda — Function.

Run conda command with environment variables set and return the json output as a julia object

Conda.prefix — Function.

Prefix for installation of the environment

Conda.python_dir — Function.

Prefix where the python command lives

Conda.rm — Function.

Uninstall a package or packages.

Conda.rm_channel — Function.

Remove a channel from the list of channels

Conda.runconda — Function.

Run conda command with environment variables set.

Conda.script_dir — Function.

Prefix for the python scripts. On UNIX, this is the same than Conda.BINDIR

Conda.search — Function.

Search packages for a string

Search a specific version of a package

Conda.update — Function.

Update all installed packages.

Conda.version — Function.

Get the exact version of a package as a VersionNumber.