Docstrings

RingBuffers.PaUtilRingBuffer — Type.

PaUtilRingBuffer(elementSizeBytes, elementCount)

Mirrors the C-side PaUtilRingBuffer struct that describes the ringbuffer state.

RingBuffers.PaUtil_FlushRingBuffer — Function.

PaUtil_FlushRingBuffer(rbuf::PaUtilRingBuffer)

Reset buffer to empty. Should only be called when buffer is NOT being read or written.

RingBuffers.PaUtil_GetRingBufferReadAvailable — Function.

PaUtil_GetRingBufferReadAvailable(rbuf::PaUtilRingBuffer)

Retrieve the number of elements available in the ring buffer for reading.

RingBuffers.PaUtil_GetRingBufferWriteAvailable — Function.

PaUtil_GetRingBufferWriteAvailable(rbuf::PaUtilRingBuffer)

Retrieve the number of elements available in the ring buffer for writing.

RingBuffers.PaUtil_InitializeRingBuffer — Function.

PaUtil_InitializeRingBuffer(rbuf, elementSizeBytes, elementCount, dataPtr)

Initialize Ring Buffer to empty state ready to have elements written to it.

Arguments

rbuf::PaUtilRingBuffer: The ring buffer.

elementSizeBytes::RingBufferSize: The size of a single data element in bytes.

elementCount::RingBufferSize: The number of elements in the buffer (must be a power of 2).

dataPtr::Ptr{Cvoid}: A pointer to a previously allocated area where the data will be maintained. It must be elementCount*elementSizeBytes long.

RingBuffers.PaUtil_ReadRingBuffer — Function.

PaUtil_ReadRingBuffer(rbuf::PaUtilRingBuffer,
                      data::Ptr{Cvoid},
                      elementCount::RingBufferSize)

Read data from the ring buffer and return the number of elements read.

RingBuffers.PaUtil_WriteRingBuffer — Function.

PaUtil_WriteRingBuffer(rbuf::PaUtilRingBuffer,
                       data::Ptr{Cvoid},
                       elementCount::RingBufferSize)

Write data to the ring buffer and return the number of elements written.

RingBuffers.RingBuffer — Type.

RingBuffer{T}(nchannels, nframes)

A lock-free ringbuffer wrapping PortAudio's implementation. The underlying representation stores multi-channel data as interleaved. The buffer will hold nframes frames.

This ring buffer can be used to pass data between tasks similar to the built-in Channel type, but it has the additional ability to pass data to and from C code running on a different thread. The C code can use the API defined in pa_ringbuffer.h. Note that this is only safe with a single-threaded reader and single-threaded writer.

RingBuffers.RingBufferSize

RingBuffers.RingBuffers

RingBuffers.__init__

RingBuffers.check_deps

RingBuffers.depsjl

RingBuffers.eval

RingBuffers.framesreadable — Function.

frameswritable(rbuf::RingBuffer)

Returns the number of frames that can be written to the ring buffer without blocking.

RingBuffers.frameswritable — Function.

frameswritable(rbuf::RingBuffer)

Returns the number of frames that can be written to the ring buffer without blocking.

RingBuffers.include

RingBuffers.libpa_ringbuffer

RingBuffers.notifyhandle — Function.

notifyhandle(rbuf::RingBuffer)

Return the AsyncCondition handle that can be used to wake up the buffer from another thread.

RingBuffers.readavailable! — Function.

readavailable!(rbuf::RingBuffer{T}, data::AbstractArray{T}[, nframes])

Read up to nframes frames from rbuf into data without blocking. If data is a Vector, it's treated as a block of memory to write the interleaved data to. If it's a Matrix, it is treated as nchannels × nframes.

Returns the number of frames read.

RingBuffers.writeavailable — Function.

writeavailable(rbuf::RingBuffer{T}, data::AbstractArray{T}[, nframes])

Write up to nframes frames to rbuf from data without blocking. If data is a Vector, it's treated as a block of memory from which to read the interleaved data. If it's a Matrix, it is treated as nchannels × nframes.

Returns the number of frames written.

RingBuffers.writeavailable!