Docstrings

GaussianRandomFields.AbstractCovarianceFunction

GaussianRandomFields.AnisotropicCovarianceStructure

GaussianRandomFields.AnisotropicExponential — Type.

AnisotropicExponential(A, σ=1)

Create an anisotropic exponential covariance structure with anisotropy matrix A and (optional) marginal standard deviation σ.

Examples

julia> A = [1 0.5; 0.5 1]
2×2 Array{Float64,2}:
 1.0  0.5
 1.0  0.5

julia> a1 = AnisotropicExponential(A)
anisotropic exponential (A=[1.0 0.5; 0.5 1.0], σ=1.0)

GaussianRandomFields.Cholesky — Type.

Cholesky <: GaussianRandomFieldGenerator

A GaussiandRandomFieldGenerator based on a Cholesky factorization of the covariance matrix.

Examples

julia> m = Matern(0.1,1.0)
Matérn (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> c = CovarianceFunction(2,m)
2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> pts1 = 0:0.02:1; pts2 = 0:0.02:1 
0.0:0.02:1.0

julia> grf = GaussianRandomField(c,Cholesky(),pts1,pts2)
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a Cholesky decomposition

julia> plot(grf) 
[...]

Note that the Cholesky factorization requires the covariance matrix to be Symmetric and positive definite. If the covariance matrix is not Symmetric, a warning will be thrown but the method will try to continue with an approximate symmetric matrix. If the covariance matrix is not positive definite, an error will be thrown. Try using the Spectral method in that case.

See also: Spectral, KarhunenLoeve, CirculantEmbedding

GaussianRandomFields.CirculantEmbedding — Type.

CirculantEmbedding <: GaussianRandomFieldGenerator

A GaussiandRandomFieldGenerator that uses FFT to compute samples of the Gaussian random field. Circulant embedding can only be applied if the points are specified on a structured grid.

Examples

julia> m = Matern(0.1,1.0)
Matérn (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> c = CovarianceFunction(2,m)
2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> pts1 = 0:0.02:1; pts2 = 0:0.02:1 
0.0:0.02:1.0

julia> grf = GaussianRandomField(c,CirculantEmbedding(),pts1,pts2)
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a circulant embedding

julia> contourf(grf)
[...]

julia> plot_eigenvalues(grf)
[...]

With appropriate ordering, the covariance matrix of a Gaussian random field is a (nested block) Toeplitz matrix. This matrix can be embedded into a larger (nested block) circulant matrix, whose eigenvalues can be rapidly computed using FFT. A difficulty here is that although the covariance matrix is positive semi-definite, its circulant extension in general is not. As a remedy, one can add so-called ghost points outside the domain of interest using the optional flag padding.

julia> m = Matern(1,1)
Matérn (λ=1.0, ν=1.0, σ=1.0, p=2.0)

julia> c = CovarianceFunction(1,m)
1d Matérn covariance function (λ=1.0, ν=1.0, σ=1.0, p=2.0)

julia> pts = range(0,stop = 1,length = 256)
0.0:0.00392156862745098:1.0

julia> grf = GaussianRandomField(c,CirculantEmbedding(),pts)
WARNING: negative eigenvalue -0.001465931113950698 detected, Gaussian random field will be approximated (ignoring all negative eigenvalues)
WARNING: increase padding if possible
Gaussian random field with 1d Matérn covariance function (λ=1.0, ν=1.0, σ=1.0, p=2.0) on a 256-point structured grid, using a circulant embedding

julia> grf = GaussianRandomField(c,CirculantEmbedding(),pts,padding=5)
Gaussian random field with 1d Matérn covariance function (λ=1.0, ν=1.0, σ=1.0, p=2.0) on a 256-point structured grid, using a circulant embedding

For faster sampling (but slower initialization), use the optional argument measure (default=true).

See also: Cholesky, Spectral, KarhunenLoeve

GaussianRandomFields.CovarianceFunction — Type.

CovarianceFunction(d, cov)

Create a covariance function in d dimensions with covariance structure cov.

Examples

julia> m = Matern(0.1,1.0)
Matérn (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> c = CovarianceFunction(2,m)
2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0)

GaussianRandomFields.CovarianceStructure

GaussianRandomFields.EOLE

GaussianRandomFields.Exponential — Type.

Exponential(λ, σ=1, p=2)

Create an exponential covariance structure with correlation length λ, (optional) marginal standard deviation σ and (optional) p-norm.

Examples

julia> e1 = Exponential(0.1)
exponential (λ=0.1, σ=1.0, p=2.0)

julia> e2 = Exponential(0.1, σ=2)
exponential (λ=0.1, σ=2.0, p=2.0)

GaussianRandomFields.FiniteElemGRF

GaussianRandomFields.FiniteElemSpectralGRF

GaussianRandomFields.GaussLegendre

GaussianRandomFields.Gaussian — Type.

Gaussian(λ, σ=1, p=2)

Create a Gaussian (squared exponential) covariance structure with correlation length λ, (optional) marginal standard deviation σ and (optional) p-norm.

Examples

julia> g1 = Gaussian(0.1)
Gaussian (λ=0.1, σ=1.0, p=2.0)

julia> g2 = Gaussian(0.1, σ=2.)
Gaussian (λ=0.1, σ=2.0, p=2.0)

GaussianRandomFields.GaussianRandomField — Type.

GaussianRandomField(mean,cov,method,pts...)
GaussianRandomField(cov,method,pts...)
GaussianRandomField(mean,cov,method,nodes,elements)
GaussianRandomField(cov,method,nodes,elements)

Compute a Gaussian random field with mean mean and covariance structure cov defined in the points pts, and computed using the method method.

Examples

julia> m = Matern(0.1,1.0)
Matérn (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> c = CovarianceFunction(2,m)
2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> pts1 = 0:0.02:1; pts2 = 0:0.02:1 
0.0:0.02:1.0

julia> mn = ones(length(pts1),length(pts2))

julia> grf = GaussianRandomField(mn,c,Cholesky(),pts1,pts2)
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a Cholesky decomposition

If no mean is specified, a zero-mean Gaussian random field is assumed.

julia> grf = GaussianRandomField(c,Cholesky(),pts1,pts2)
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a Cholesky decomposition

The Gaussian random field sampler method can be Cholesky(), Spectral(), KarhunenLoeve(n) (where n is the number of terms in the expansion), or CirculantEmbedding(). The dimension of the points must match the dimension of the covariance function cov. The points can be specified as arguments of type AbstractVector, in which case a tensor (Kronecker) product is assumed, or as a Finite Element mesh with node table nodes and element table elements.

julia> grf = GaussianRandomField(c,KarhunenLoeve(500),pts1,pts2)
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a KL expansion with 500 terms

For separable Gaussian random fields with a KL expansion, provide a SeparableCovarianceFunction. Note that the number of terms n in KarhunenLoeve(n) refers to the total number of terms.

julia> e1 = Exponential(0.1); e2 = Exponential(0.01);

julia> scov = SeparableCovarianceFunction(e1,e2)
2d separable covariance function [ exponential (λ=0.1, σ=1.0, p=2.0), exponential (λ=0.01, σ=1.0, p=2.0) ]

julia> grf = GaussianRandomField(scov,KarhunenLoeve(500),pts1,pts2)
Gaussian random field with 2d separable covariance function [ exponential (λ=0.1, σ=1.0, p=2.0), exponential (λ=0.01, σ=1.0, p=2.0) ] on a 51x51 structured grid, using a KL expansion with 500 terms

julia> plot_eigenfunction(grf,3); show()
[...]

Also anisotropic Gaussian random fields can be computed. For example, the anisotropic exponential covariance function needs a positive definite matrix A. The size of the off-diagonal elements of A determine the degree of anisotropy.

julia> a = AnisotropicExponential([1 0.8;0.8 1])
anisotropic exponential (A=[1.0 0.8; 0.8 1.0], σ=1.0)

julia> acov = CovarianceFunction(2,a)
2d anisotropic covariance function exponential (A=[1.0 0.8; 0.8 1.0], σ=1.0)

julia> pts = range(0,stop = 10,length = 128)
0.0:0.07874015748031496:10.0

julia> grf = GaussianRandomField(acov,CirculantEmbedding(),pts,pts)
WARNING: negative eigenvalue -1.2245106889248049e-18 detected, Gaussian random field will be approximated (ignoring all negative eigenvalues)
WARNING: increase padding if possible
Gaussian random field with 2d anisotropic covariance function exponential (A=[1.0 0.8; 0.8 1.0], σ=1.0) on a 128x128 structured grid, using a circulant embedding

julia> contourf(grf)
[...]

Samples from the random field can be computed using the sample function.

julia> sample(grf)
[...]

See also: Cholesky, Spectral, KarhunenLoeve, CirculantEmbedding, sample

GaussianRandomFields.GaussianRandomFieldGenerator

GaussianRandomFields.GaussianRandomFields

GaussianRandomFields.IsotropicCovarianceStructure

GaussianRandomFields.KarhunenLoeve — Type.

KarhunenLoeve{n} <: GaussianRandomFieldGenerator

A GaussiandRandomFieldGenerator using a Karhunen-Loève (KL) expansion with n terms.

Examples

julia> m = Matern(0.1,1.0)
Matérn (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> c = CovarianceFunction(2,m)
2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> pts1 = 0:0.02:1; pts2 = 0:0.02:1 
0.0:0.02:1.0

julia> grf = GaussianRandomField(c,KarhunenLoeve(300),pts1,pts2)
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a KL expansion with 300 terms

julia> plot_eigenvalues(grf) # plot the eigenvalue decay
[...]

julia> plot_eigenfunction(grf,4) # plots the fourth eigenfunction
[...]

The more terms are retained in the expansion, the better the approximation will be.

julia> nterms = [1 2 5 10 20 50 100 200 500 1000]
1×10 Array{Int64,2}:
 1  2  5  10  20  50  100  200  500  1000

julia> for n in nterms
       grf = GaussianRandomField(c,KarhunenLoeve(n),pts1,pts2)
       @show rel_error(grf)
       end
rel_error(grf) = 0.7499982529722711
rel_error(grf) = 0.49999825591379987
rel_error(grf) = 0.4425751861338164
rel_error(grf) = 0.35789475278408045
rel_error(grf) = 0.16805079842673853
rel_error(grf) = 0.11187098338277579
rel_error(grf) = 0.05130466343704787
rel_error(grf) = 0.017343327476498027
rel_error(grf) = 0.0034278579378175245
rel_error(grf) = 0.0007216777927243623

The KL expansion is computed using the Nystrom method. Optional argument are the quadrature rule quad and the number of quadrature points in each direction nq. Possible values for quad are GaussLegendre(), Midpoint(), Simpson(), Trapezoidal and EOLE() (default). The total number of quadrature points must be larger than or equal to the requested number of terms. Default is ceil(n^(1/d)). This value is best left untouched.

julia> grf = GaussianRandomField(c,KarhunenLoeve(300),pts1,pts2,quad=GaussLegendre())
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a KL expansion with 300 terms

julia> grf = GaussianRandomField(c,KarhunenLoeve(300),pts1,pts2,quad=Simpson(),nq=20)
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a KL expansion with 300 terms

See also: Cholesky, Spectral, CirculantEmbedding

GaussianRandomFields.LSHAPE_P

GaussianRandomFields.LSHAPE_T

GaussianRandomFields.Lshape

GaussianRandomFields.Matern — Type.

Matern(λ, ν, σ=1, p=2)

Create a Matérn covariance structure with correlation length λ, smoothness ν, (optional) marginal standard deviation σ and (optional) p-norm.

Examples

julia> m1 = Matern(0.1, 1.0)
Matérn (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> m2 = Matern(0.1, 1.0, σ=2.0)
Matérn (λ=0.1, ν=1.0, σ=2.0, p=2.0)

GaussianRandomFields.Midpoint

GaussianRandomFields.OneDimCov

GaussianRandomFields.OneDimGRF

GaussianRandomFields.OneDimSepCov

GaussianRandomFields.OneDimSpectralGRF

GaussianRandomFields.OneDimSpectralSepGRF

GaussianRandomFields.QuadratureRule

GaussianRandomFields.STAR_P

GaussianRandomFields.STAR_T

GaussianRandomFields.SeparableCovarianceFunction — Type.

SeparableCovarianceFunction(cov...)

Create a separable covariance function in `length(cov)` dimensions for the covariance structures `cov`. Usefull for defining anisotropic covariance functions, or if the usual KL expansion is too expensive.

Examples

julia> e = Exponential(0.1)
exponential (λ=0.1, σ=1.0, p=2.0)

julia> m = Matern(0.01,1.0)
Matérn (λ=0.01, ν=1.0, σ=1.0, p=2.0)

julia> c = SeparableCovarianceFunction([e,m])
2d separable covariance function [ exponential (λ=0.1, σ=1.0, p=2.0), Matérn (λ=0.01, ν=1.0, σ=1.0, p=2.0) ]

See also: CovarianceFunction, KarhunenLoeve

GaussianRandomFields.Simpson

GaussianRandomFields.Spectral — Type.

Spectral <: GaussianRandomFieldGenerator

A GaussiandRandomFieldGenerator based on a spectral (eigenvalue) decomposition of the covariance matrix.

Examples

julia> m = Matern(0.1,1.0)
Matérn (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> c = CovarianceFunction(2,m)
2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> pts1 = 0:0.02:1; pts2 = 0:0.02:1 
0.0:0.02:1.0

julia> grf = GaussianRandomField(c,Spectral(),pts1,pts2)
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a Spectral decomposition

julia> plot(grf) 
[...]

This is also useful when computing Gaussian random fields on a Finite Element mesh using a truncated KL expansion. Here's an example that computes the first 10 eigenfunctions on an L-shaped domain.

julia> p, t = Lshape();

julia> grf = grf = GaussianRandomField(CovarianceFunction(2,Matern(0.2,1.0)),Spectral(),p,t,n=10)
Gaussian random field with 2d Matérn covariance function (λ=0.2, ν=1.0, σ=1.0, p=2.0) on a mesh with 998 points and 1861 elements, using a spectral decomposition

julia> tricontourf(p[:,1],p[:,2],grf.data.eigenfunc[:,1],triangles=t-1,cmap=get_cmap("viridis"))
[...]

See also: Cholesky, KarhunenLoeve, CirculantEmbedding

GaussianRandomFields.SpectralData

GaussianRandomFields.SquaredExponential — Type.

SquaredExponential(λ, σ=1, p=2)

Create a squared exponential (Gaussian) covariance structure with correlation length λ, (optional) marginal standard deviation σ and (optional) p-norm.

Examples

julia> s1 = SquaredExponential(0.1)
Gaussian (λ=0.1, σ=1.0, p=2.0)

julia> s2 = SquaredExponential(0.1, σ=2.)
Gaussian (λ=0.1, σ=2.0, p=2.0)

GaussianRandomFields.ThreeDimCov

GaussianRandomFields.ThreeDimGRF

GaussianRandomFields.ThreeDimSepCov

GaussianRandomFields.ThreeDimSpectralGRF

GaussianRandomFields.ThreeDimSpectralSepGRF

GaussianRandomFields.Trapezoidal

GaussianRandomFields.TwoDimCov

GaussianRandomFields.TwoDimGRF

GaussianRandomFields.TwoDimSepCov

GaussianRandomFields.TwoDimSpectralGRF

GaussianRandomFields.TwoDimSpectralSepGRF

GaussianRandomFields._GaussianRandomField

GaussianRandomFields._plot_eigenfunction

GaussianRandomFields._sample

GaussianRandomFields.apply

GaussianRandomFields.apply_non_symmetric!

GaussianRandomFields.apply_symmetric!

GaussianRandomFields.bisect_root

GaussianRandomFields.circulant_eigvals — Function.

circulant_eigvals(cov::CovarianceStructure, pts::NTuple{N,AbstractRange},
                  dims::Dims{N}) where {N}

Compute eigenvalues of circulant embedding with dimensions dims of covariance matrix of points pts with covariance function cov.

GaussianRandomFields.circulant_minsize — Function.

circulant_minsize(cov::CovarianceStructure, pts::AbstractRange,
                  minpadding::Int, factors::Vector{Int})

Return size of minimum circulant embedding of covariance matrix of points pts with covariance function cov and minimum padding minpadding that can be written as product of integers in factors.

Typically factors is chosen to be [2] or [2, 3, 5, 7] to speed up FFT computations.

GaussianRandomFields.compute_analytic

GaussianRandomFields.compute_centers

GaussianRandomFields.eval

GaussianRandomFields.extrapolate

GaussianRandomFields.findroots

GaussianRandomFields.generator

GaussianRandomFields.get_nodes_and_weights

GaussianRandomFields.include

GaussianRandomFields.middle

GaussianRandomFields.plot_covariance_matrix

GaussianRandomFields.plot_eigenfunction

GaussianRandomFields.plot_eigenvalues

GaussianRandomFields.randdim

GaussianRandomFields.rel_error

GaussianRandomFields.sample — Function.

sample(grf)
sample(grf, xi=randn(randdim(grf)))

Take a sample from the Gaussian Random Field grf using the (optional) random numbers xi. The vectorxi must have appropriate length. The default value is randn(randdim(grf)).

Examples

julia> m = Matern(0.1,1.0)
Matérn (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> c = CovarianceFunction(2,m)
2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0)

julia> pts1 = 0:0.02:1; pts2 = 0:0.02:1 
0.0:0.02:1.0

julia> grf = GaussianRandomField(c,KarhunenLoeve(300),pts1,pts2)
Gaussian random field with 2d Matérn covariance function (λ=0.1, ν=1.0, σ=1.0, p=2.0) on a 51x51 structured grid, using a KL expansion with 300 terms

julia> sample(grf)
[...]

julia> sample(grf,xi=2*rand(randdim(grf))-1)
[...]

GaussianRandomFields.shape

GaussianRandomFields.shortname

GaussianRandomFields.showparams

GaussianRandomFields.showpoints

GaussianRandomFields.slice

GaussianRandomFields.star