Docstrings

BEAST.@discretise — Macro.

discr(eq, pairs...)

This macro provides syntactical sugar for the definition of a discretisation of a varational formulation. Given a variational equation EQ: Find j ∈ X such that for all k ∈ Y a(k,j) = f(k) can be discretised by stating:

eq = @discretise EQ j∈X k∈Y

BEAST.@hilbertspace

BEAST.Variational.@varform — Macro.

@varform <form-definition>

The Julia form compiler uses the Julia parser and meta-programming based traversal of the AST to create a structure containing all information required for the description of a variational formulation from an Expr that follows closely widespread mathematical convention.

E.g:

EFIE = @varform T[k,j] = e[k]
MFIE = @varform 0.5*I[k,j] + K[k,j] = h[k]
PMCH = @varform M[k,j] - η*T[k,m] + 1/η*T[l,j] + M[l,m] = e[k] + h[l]

BEAST.ADIterator

BEAST.AbstractOperator

BEAST.AbstractSpace

BEAST.AbstractTimeBasisFunction

BEAST.AssemblyData

BEAST.AssemblyDataEl

BEAST.BDMBasis

BEAST.BDMRefSpace

BEAST.BEAST

BEAST.Bernstein

BEAST.Continuity

BEAST.CrossTraceMW

BEAST.CurlSingleLayerDP3D

BEAST.DiagonalizedMatrix

BEAST.DiracBoundary

BEAST.DirectProductSpace

BEAST.DiscreteEquation

BEAST.DoubleLayer

BEAST.DoubleLayerRotatedMW3D

BEAST.DoubleLayerTransposed

BEAST.DoubleQuadStrategy

BEAST.EmptyRP

BEAST.ErrorFunction

BEAST.Functional

BEAST.GMRESSolver

BEAST.Gaussian

BEAST.HH3DDoubleLayer

BEAST.HH3DDoubleLayerTDBIO

BEAST.HH3DDoubleLayerTransposed

BEAST.HH3DHyperSingularFDBIO — Type.

∫_Γ dx ∫_Γ dy \left(α G g(x) n_x ⋅ n_y f(y) + β G \mbox{curl} g(x) ⋅ \mbox{curl} f(y) \right)

with $G(x,y) = \frac{e^{-γ |x-y|}}{4 π |x-y|}$

BEAST.HH3DPlaneWave

BEAST.HH3DSingleLayerFDBIO — Type.

\[a(u,v) = α ∬_{Γ×Γ} u(x) G_{γ}(|x-y|) v(y)\]

with $G_{γ}(r) = \frac{e^{-γr}}{4πr}$.

BEAST.HH3DSingleLayerReg

BEAST.HH3DSingleLayerSng

BEAST.HH3DSingleLayerTDBIO

BEAST.Helmholtz3D

BEAST.Helmholtz3DOp

BEAST.Helmholtz3DOpReg

BEAST.HelmholtzOperator2D

BEAST.HyperSingular

BEAST.Identity

BEAST.IntegralOperator

BEAST.KernelValsHelmholtz2D

BEAST.KernelValsMaxwell3D

BEAST.LagrangeBasis

BEAST.LagrangeRefSpace

BEAST.LinearCombinationOfOperators

BEAST.LinearSpace

BEAST.LocalOperator

BEAST.MWDL3DGen

BEAST.MWDL3DIntegrand

BEAST.MWDL3DIntegrand2

BEAST.MWDoubleLayer3D

BEAST.MWDoubleLayer3DReg

BEAST.MWDoubleLayer3DSng

BEAST.MWDoubleLayerTDIO

BEAST.MWFarField3D — Type.

Operator to compute the far field of a current distribution. In particular, given the current distribution $j$ this operator allows for the computation of

\[A j = n × ∫_Γ e^{γ ̂x ⋅ y} dy\]

where $̂x$ is the unit vector in the direction of observation. Note that the assembly routing expects the observation directions to be normalised by the caller.

BEAST.MWFarField3DTD — Type.

Operator to compute the far field of a current distribution in the time domain. In particular, given the current distribution $j$ this operator allows for the computation of

\[R = ̂x ⋅ y ffd = n × ∫_Γ j(r', t - R/c} dy\]

where $̂x$ is the unit vector in the direction of observation. Note that the assembly routing expects the observation directions to be normalised by the caller.

BEAST.MWHyperSingular

BEAST.MWOperator3D

BEAST.MWSL3DGen

BEAST.MWSL3DIntegrand

BEAST.MWSL3DIntegrand2

BEAST.MWSingleLayer3D

BEAST.MWSingleLayer3DReg

BEAST.MWSingleLayer3DSng

BEAST.MWSingleLayerField3D — Type.

MWSingleLayerField3D(wavenumber = error())

Create the single layer near field operator, for use with potential.

BEAST.MWSingleLayerPotential3D

BEAST.MWSingleLayerTDIO

BEAST.MWWeaklySingular

BEAST.MatrixConvolution

BEAST.Maxwell3D

BEAST.MaxwellOperator3D

BEAST.MaxwellOperator3DReg

BEAST.MonomialBasis

BEAST.MultiQuadStrategy

BEAST.NCross

BEAST.NDBasis

BEAST.NDRefSpace

BEAST.NDotTrace

BEAST.NitscheHH3

BEAST.NormalDerivative

BEAST.NormalVector

BEAST.NumQuadStrategy

BEAST.Operator — Type.

Atomic operator: one that assemblechunk can deal with

BEAST.PieceWisePolynomial

BEAST.PlaneWaveDirichlet

BEAST.PlaneWaveHH3DTD

BEAST.PlaneWaveMW

BEAST.PlaneWaveMWTD

BEAST.PlaneWaveNeumann

BEAST.Polynomial

BEAST.RTBasis

BEAST.RTRefSpace

BEAST.RefSpace

BEAST.RetardedPotential

BEAST.RungeKuttaConvolutionQuadrature — Type.

RungeKuttaConvolutionQuadrature{T,N,NN}

T: the value type of the basis function. N: the number of stages. NN: N*N.

Performs a convolution quadrature on a laplaceKernel to represent an operator in time domain using an implicit Runge-Kutta method.

laplaceKernel: function of the Laplace variable s that returns an IntegralOperator. A, b: Coefficient matrix and vectors from the Butcher tableau. Δt: time step. zTransformedTermCount: Number of terms in the inverse Z-transform. contourRadius: radius of circle used as integration contour for the inverse Z-transform.

BEAST.ScalarTrace

BEAST.Shape

BEAST.SingleLayer

BEAST.SingleLayerTrace — Type.

Describe a single layer operator from the surface to a line.

\[<v, Su> = ∫_γ dx v(x) ∫_Γ dy rac{e^{-ikR}}{4πR} u(y)\]

BEAST.SingleQuadStrategy

BEAST.SingleQuadStrategy2

BEAST.SingularityExtractionStrategy

BEAST.Space

BEAST.SpaceTimeBasis

BEAST.SparseND

BEAST.StagedTimeStep — Type.

StagedTimeStep{T,N}

T: the value type of the basis function. N: the number of stages. It corresponds to a time-space basis function where each time step has intermediary stages given by the vertor c in a Butcher tableau (A,b,c)

BEAST.TDFunctional

BEAST.TangTraceMW

BEAST.TemporalDifferentiation

BEAST.TensorOperator

BEAST.TimeBasisDelta

BEAST.TimeBasisDeltaShifted — Type.

TimeBasisDeltaShifted{T}

Represents a TimeBasisDelta{T} retarded by a fraction of the time step.

BEAST.TimeBasisFunction — Type.

TimeBasisFunction{N,D}

T: the value type of the time basis function N: the number of intervals in the support (this included the semi infinite interval stretching to +∞) D1: the degree of the TBF restricted to each of the intervals plus one

BEAST.TransposedOperator

BEAST.ValDiv

BEAST.Variational

BEAST.WiltonInts84Strat

BEAST.WiltonSEStrategy

BEAST._legendre

BEAST.add!

BEAST.allocatestorage

BEAST.assemble — Function.

assemble(fn, tfs)

Assemble the vector of test coefficients corresponding to functional fn and test functions tfs.

BEAST.assemble!

BEAST.assemble_local_matched!

BEAST.assemble_local_mixed! — Function.

assemble_local_mixed(biop::LocalOperator, tfs, bfs)

For use when basis and test functions are defined on different meshes

BEAST.assemble_st!

BEAST.assembleblock

BEAST.assembleblock!

BEAST.assembleblock_body!

BEAST.assembleblock_primer

BEAST.assemblechunk! — Function.

assemblechunk!(biop::IntegralOperator, tfs, bfs, store)

Computes the matrix of operator biop wrt the finite element spaces tfs and bfs

BEAST.assemblechunk_body!

BEAST.assemblechunk_body_nested_meshes!

BEAST.assemblecol

BEAST.assemblecol!

BEAST.assemblecol_body!

BEAST.assemblerow

BEAST.assemblerow!

BEAST.assemblerow_body!

BEAST.assembly

BEAST.assemblydata — Function.

charts, admap = assemblydata(basis)

Given a Basis this function returns a data structure containing the information required for matrix assemble. More precise the following expressions are valid for the returned object ad:

ad[c,r,i].globalindex
ad[c,r,i].coefficient

Here, c and r are indices in the iterable set of geometric elements and the set of local shape functions on each element. i ranges from 1 to the maximum number of basis functions local shape function r on element r contributes to.

For a triplet (c,r,i), globalindex is the index in the Basis of the i-th basis function that has a contribution from local shape function r on element r. coefficient is the coefficient of that contribution in the linear combination defining that basis function in terms of local shape function.

Note: the indices c correspond to the position of the corresponding element whilst iterating over geometry(basis).

BEAST.basisfunction

BEAST.blockassembler — Function.

blockassembler(operator, test_space, trial_space) -> assembler

Return a callable object for the creation of blocks within a BEM matrix.

This function performs all tasks common to the assembly of several blocks within a single boundary element matrix. The return value can be used to generate blocks by calling it as follows:

assembler(I,J,storefn)

where I and J are arrays of indices in test_space and trial_space, respectively, corresponding to the rows and columns of the desired block.

Note that the block will be constructed in compressed form, i.e. the rows and columns of the store that are written into are the positions within I and J (as opposed to the positions within 1:numfunctions(test_space) and 1:numfunctions(trial_space)). In particular the size of the constructed block will be (length(I), length(J)).

This last property allows the assembly of permutations of the BEM matrix by supplying for I and J permutations of 1:numfunctions(test_space) and 1:numfunctions(trial_space).

BEAST.brezzidouglasmarini

BEAST.buffachristiansen — Function.

buffachristiansen(Γ, γ)

Construct the set of Buffa-Christiansen functions subject to mesh Γ and only enforcing zero normal components on ∂Γ ∖ γ.

BEAST.buildhalfbc — Function.

buildhalfbc(fine, supp::Array{SVector{3,Int},1}, v, p)

BEAST.butcher_tableau_radau_2stages — Function.

butcher_tableau_radau_2stages()

Returns (A,b,c) corresponding to the Butcher tableau for the 2 stage Radau IIA scheme.

BEAST.butcher_tableau_radau_3stages — Function.

butcher_tableau_radau_3stages()

Returns (A,b,c) corresponding to the Butcher tableau for the 3 stage Radau IIA scheme.

BEAST.cellcellinteractions!

BEAST.cellinteractions

BEAST.celltestvalues

BEAST.collapse_shapes

BEAST.companion

BEAST.convolve

BEAST.coordtype

BEAST.creategaussian

BEAST.cross

BEAST.curl — Function.

curl(X)

Compute the curl of a finite element basis. The resulting set of functions might be linearly dependent because of the kernel of the curl operator.

BEAST.degree

BEAST.derive

BEAST.diagonalizedmatrix

BEAST.discretise

BEAST.divergence — Function.

divergence(x)

Compute the divergence of a finite element space.

BEAST.dot

BEAST.duallagrangec0d1 — Function.

duallagrangec0d1(originalmesh, refinedmesh)

It is the user responsibility to provide two meshes representing the same object. The second mesh needs to be obtained using "barycentric_refinement(originalmesh)". This basis function creats the dual Lagrange basis function and return an object that contains array of shapes [fns] It also return a gemoetry containing the refined mesh.

BEAST.duallagrangecxd0 — Function.

duallagrangecxd0(mesh, jct) -> basis

Build dual Lagrange piecewise constant elements. Boundary nodes are only considered if they are in the interior of jct.

BEAST.elements — Function.

elements(geo)

Create an iterable collection of the elements stored in geo. The order in which this collection produces the elements determines the index used for lookup in the data structures returned by assemblydata and quaddata.

BEAST.elementstree

BEAST.eval

BEAST.evaluate

BEAST.facecurrents — Function.

fcr, geo = facecurrents(coeffs, basis)

Compute the value of the function with the given collection of coeffient in the provided basis in all the centroids of the mesh underlying the basis. The mesh is returned together with the currents.

BEAST.faces

BEAST.farfieldlocal!

BEAST.fouriertransform

BEAST.geometry — Function.

geometry(basis)

Returns an iterable collection of geometric elements on which the functions in basis are defined. The order the elements are encountered needs correspond to the element indices used in the data structure returned by assemblydata.

BEAST.get_scatter_parameters

BEAST.getcommonedge — Function.

getcommonedge(cell1, cell2) -> e1, e2, edge

Returns in edge the common vertices of cell1 and cell2. e1 contains the index of the vertex of cell1 opposite to this common edge, and with a plus or minus sign depending on whether the orientation of the common edge is along or against the internal orientation of cell1. Similar for e2.

BEAST.getindex_rtg — Function.

getindex_rtg(RT::RTBasis)

Returns the indices of the global half RWGs present in RT. RT is typically gotten from rt_ports

BEAST.gmres

BEAST.include

BEAST.index_actives

BEAST.innerintegrals!

BEAST.instantiate_charts

BEAST.integrand

BEAST.integrate

BEAST.interior_and_junction_vertices

BEAST.inverse_z_transform — Function.

inverse_z_transform(k, rho, N, X)

Returns the k-th term of the inverse z-transform. X is an array of the z-transform evaluated in the points z=rhoexp(2impin/N) for n in 0:(N-1).

BEAST.isclosed

BEAST.kernel_in_bary

BEAST.kernelvals

BEAST.lagdimension — Function.

The dimension of the space of Lagrange shape functions of degree d over a simplex of dimension n is binom(n+d,d) == binom(n+d,n)

BEAST.lagrangec0d1 — Function.

lagrangec0d1(mesh; dirichlet=[true|false]) -> basis

Build lagrangec0d1 elements, including (dirichlet=false) or excluding (dirichlet=true) those attached to boundary vertices.

BEAST.lagrangec0d1_dirichlet — Function.

lagrangec0d1(mesh[, bnd])

Construct the basis of continuous, piecewise linear basis functions subordinate to mesh mesh. Basis functions are constructed at vertices in the interionr of the mesh and on the closure of 'bnd'. In particular, leaving out the second argument creates a finite element space subject to homogeneous Dirichlet boundary conditions.

BEAST.lagrangecxd0

BEAST.laplace_to_z — Function.

laplace_to_z(rho, n, N, dt, A, b)

Returns the complex matrix valued Laplace variable s that correspond to the variable z = rhoexp(2impin/N) for a given Butcher tableau (A,b,c) and a time step dt.

BEAST.localoperator

BEAST.localoperator2

BEAST.make_celltonum

BEAST.marchonintime — Function.

marchonintime(W0,Z,B,I)

Solve by marching-on-in-time the causal convolution problem defined by (W0,Z,B) up to timestep I. Here, Z is an array of order 3 that contains a discretisation of a time translation invariant retarded potential operator. W0 is the inverse of the slice Z[:,:,1].

BEAST.momintegrals! — Function.

regularcellcellinteractions!(biop, tshs, bshs, tcell, bcell, interactions, strat)

Function for the computation of moment integrals using simple double quadrature.

BEAST.momintegrals_nested!

BEAST.move_after! — Function.

Move the s-th element right after the d-th

BEAST.move_before!

BEAST.n

BEAST.nedelec

BEAST.ntrace — Function.

ntrace(X::Space, γ::Mesh)

Compute the normal trace of basis X on mesh γ. γ is assumed to be part of the boundary of geometry(X).

ntrace(refspace, element, localindex, face)

Compute the normal trace of all local shape functions on elements belonging to refspace on face. This function returns a matrix expressing the traces of local shape functions in refspace as linear combinations of functions in the local trace space. Cf. restrict. localindex is the index of face in the enumeration of faces of elements. In many special cases knowing this index allows for highly optimised implementations.

BEAST.numfunctions — Function.

numfunctions(r)

Return the number of functions in a Space or RefSpace.

BEAST.numintervals

BEAST.operator

BEAST.planewave

BEAST.planewavemw3d — Function.

planewavemw3d(;direction, polarization, wavenumber[, amplitude=1])

Create a plane wave solution to Maxwell's equations.

BEAST.portcells — Function.

portcells(Γ::Mesh, γ::Mesh)

returns an array containing cell pairs of mesh Γ around a boundary edge that overlaps with mesh γ

BEAST.positions

BEAST.potential

BEAST.potential!

BEAST.qh — Function.

Q = qd(T,dh,::Val{N})

Q[k] is the factor in front resulting from differentiating t^(k-1) dh times.

BEAST.qrdf

BEAST.qrib

BEAST.qrss

BEAST.quaddata — Function.

quaddata(operator, test_refspace, trial_refspace, test_elements, trial_elements)

Returns an object cashing data required for the computation of boundary element interactions. It is up to the client programmer to decide what (if any) data is cached. For double numberical quadrature, storing the integration points for example can significantly speed up matrix assembly.

operator is an integration kernel.

test_refspace and trial_refspace are reference space objects. quadata

is typically overloaded on the type of these local spaces of shape functions. (See the implementation in maxwell.jl for an example).

test_elements and trial_elements are iterable collections of the geometric

elements on which the finite element space are defined. These are provided to allow computation of the actual integrations points - as opposed to only their coordinates.

BEAST.quadrule — Function.

quadrule(operator,test_refspace,trial_refspace,p,test_element,q_trial_element, qd)

Returns an object that contains all the dynamic (runtime) information that defines the integration strategy that will be used by momintegrals! to compute the interactions between the local test/trial functions defined on the specified geometric elements. The indices p and q refer to the position of the test and trial elements as encountered during iteration over the output of geometry.

The last argument qd provides access to all precomputed data required for quadrature. For example it might be desirable to precompute all the quadrature points for all possible numerical quadrature schemes that can potentially be required during matrix assembly. This makes sense, since the number of point is order N (where N is the number of faces) but these points will appear in N^2 computations. Precomputation requires some extra memory but can save a lot on computation time.

BEAST.raowiltonglisson — Function.

raviartthomas(mesh, cellpairs::Array{Int,2})

Constructs the RT basis on the input mesh. The i-th RT basis function will represent a current distribution flowing from cell cellpairs[1,i] to cellpairs[2,i] on the mesh.

Returns an object of type RTBasis, which comprises both the mesh and pairs of Shape objects which corresponds to the cell pairs, containing the necsessary coefficients and indices to compute the exact basis functions when required by the solver.

raviartthomas(mesh)

Conducts pre-processing on the input mesh by extracting the cell edges, cell pairs and indices required to construct the RT basis on the mesh.

Calls raviartthomas(mesh::Mesh, cellpairs::Array{Int,2}), which constructs the RT basis on the mesh, using the cell pairs identified.

Returns the RT basis object.

raviartthomas(Γ, γ)

Constructs the RT space relative to boundary γ of an open surface, only selecting cell pairs whose common edge does not lie on γ . (This prevents the calculation of physically-impossible surface currents, such as those flowing 'off the edge' of a surface.)

Calls raviartthomas(Γ, duals) which constructs the RT basis on the mesh, using the cell pairs identified.

BEAST.raviartthomas — Function.

raviartthomas(mesh, cellpairs::Array{Int,2})

Constructs the RT basis on the input mesh. The i-th RT basis function will represent a current distribution flowing from cell cellpairs[1,i] to cellpairs[2,i] on the mesh.

Returns an object of type RTBasis, which comprises both the mesh and pairs of Shape objects which corresponds to the cell pairs, containing the necsessary coefficients and indices to compute the exact basis functions when required by the solver.

raviartthomas(mesh)

Conducts pre-processing on the input mesh by extracting the cell edges, cell pairs and indices required to construct the RT basis on the mesh.

Calls raviartthomas(mesh::Mesh, cellpairs::Array{Int,2}), which constructs the RT basis on the mesh, using the cell pairs identified.

Returns the RT basis object.

raviartthomas(Γ, γ)

Constructs the RT space relative to boundary γ of an open surface, only selecting cell pairs whose common edge does not lie on γ . (This prevents the calculation of physically-impossible surface currents, such as those flowing 'off the edge' of a surface.)

Calls raviartthomas(Γ, duals) which constructs the RT basis on the mesh, using the cell pairs identified.

BEAST.real_inverse_z_transform — Function.

real_inverse_z_transform(k, rho, N, X)

Returns the k-th term of the inverse z-transform. It is assumed that X[n+1] = conj(X[N-n]) for each n in 1:(N-1) so that Nmax = N/2+1 or (N+1)/2 (resp. if N%2==0 or N%2==1) terms are used in X X is an array of the z-transform evaluated in the points z=rhoexp(2impin/N) for n in 0:(Nmax-1).

BEAST.refspace — Function.

refspace(basis)

Returns the ReferenceSpace of local shape functions on which the basis is built.

BEAST.regularpart

BEAST.regularpart_quadrule

BEAST.relorientation

BEAST.restrict — Function.

restrict(refspace, element1, element2)

Computes the restriction of a set of local shape functions on element1 as linear combinations of the set of local shape functions on element2. More precisely restrict returns an NxM matrix P such that the i-th local shape $g_i$ function on element2 can be written as:

$g_i = sum_{j=1}^{M} P_{ij} f_j$

BEAST.rhs

BEAST.rt_cedge — Function.

rt_cedge(cps::Array{Int,2}, weight)

Computes single basis function with equally distributed constant current leaving or entering port defined by cellpairs cps. weight defines the total current over the port and its direction (+ve = out, -ve = in)

BEAST.rt_ports — Function.

rt_ports(Γ::Mesh, γ::Mesh ...)

Constructs the RT space on Γ, relative to boundary pairs in γ. γ expects any number of pairs-of-ports as arguments and accepts tuples, arrays, vectors etc. e.g rt_ports(Γ, a, b ...); where a = [γ₁ γ₂], b = (γ₃,γ₄) etc. rt_ports with no pair of ports supplied i.e rt_ports(Γ) reduces to the raviartthomas(Γ) function. The RT space ensures current continuity in each pair of ports. i.e. current leaving mesh Γ through γ₁ is accounted for in γ₂.

Returns the RT basis object.

BEAST.rt_vedge — Function.

rt_vedge(cps::Array{Int,2}, weight)

Computes n-1 basis function with oscillating current in(leaving and entering) pairs of half triangles defined over port specified by cellpairs cps. weight defines the magnitude of individual current in and out the half triangles, and it's polarity simply defines whether to start with in or out

BEAST.scalartype — Function.

scalartype(x)

The scalar field over which the values of a global or local basis function, or an operator are defined. This should always be a scalar type, even if the basis or operator takes on values in a vector or tensor space. This data type is used to determine the eltype of assembled discrete operators.

BEAST.shapevals

BEAST.singleduallagd0 — Function.

singleduallagd0(fine, F, v)

Build a single dual constant Lagrange element a mesh fine. F contains the indices to cells in the support and v is the index in the vertex list of the defining vertex.

BEAST.singularpart

BEAST.solve — Function.

Solves a variational equation by simply creating the full system matrix and calling a traditional lu decomposition.

BEAST.sortneighbors

BEAST.spatialbasis

BEAST.strace

BEAST.subReferenceSpace

BEAST.subdBasis

BEAST.subdsurface

BEAST.subset

BEAST.substitute

BEAST.sysmatrix

BEAST.td_solve

BEAST.temporalassemblydata

BEAST.temporalbasis

BEAST.testfunc1

BEAST.timebasisc0d1 — Function.

timebasisc0d1(type, timestep, numfunctions)

Build the space of continuous, piecewise linear time basis functions. The DoFs are the time steps. numfunctions basis functions will be built in total.

BEAST.timebasiscxd0 — Function.

timebasiscxd0(timestep, numfunctions, T::Type=Float64)

Create a temporal basis based on shifted copies of the nodal continuous, piecewise linear interpolant.

BEAST.timebasisdelta

BEAST.timebasisshiftedlagrange

BEAST.timebasisspline2 — Function.

timebasisspline2(timestep, numfunctions, T::Type=Float64)

Create a temporal basis based on shifted copies of the quadratic spline. The spline is the convolution of a cxd0 and a c0d1 basis function.

BEAST.timeintegrals!

BEAST.timequadrule

BEAST.timestep

BEAST.tmRoR

BEAST.tmRoRf

BEAST.valuetype

BEAST.x̂

BEAST.×

BEAST.ŷ

BEAST.ẑ

BEAST.∂n

BEAST.⊗