Docstrings

Signals.DropRepeats

Signals.Filter

Signals.Merge

Signals.PullAction

Signals.PullType

Signals.RecursionFree

Signals.Signal — Type.

S = Signal(val; strict_push = false)

Create a source Signal with initial value val, setting strict_push to true guarantees that every push to this Signal will be carried out independently. Otherwise if updates occur faster than what the eventloop can process, then only the last value before the eventloop kicks in will be used(default).

S = Signal(f,args...; v0 = nothing)

Create a derived Signal whos value is f(args...), args can be of any type, Signal args get replaced by their value before calling f(args...). reads best with with do notation(see example).if v0 is not nothing then f(args...) will not be called directly after Signal creation instead the Signal will be initialized to have value v0.

Syntax

S[] = val

Set the value of S to val without propogating the change to the rest of the signal graph, usefull in pull based paradigm.

S()

pull! Signal, pulling any changes in the Signal graph that affects S.

S(val)

Set the value of S to val and pushes the changes along the Signal graph.

S[]

Get the current value stored in S without pulling changes from the graph.

Examples

julia> A = Signal(1) # source Signal
Signal
value: 1

julia> B = 2 # non-Signal input
2

julia> C = Signal(A, B) do a, b # derived Signal
           a + b
       end

Signal
value: 3

julia> A[] = 10 # set value without propogation
10
julia> C[] # reads current value
3
julia> C() # pull latest changes from the Signal graph
12
julia> A(100) # set value to a signal and propogate this change
100
julia> C[]
102

Signals.SignalData

Signals.SignalException

Signals.Signals

Signals.StandardPull

Signals.Stateless

Signals.Throttle

Signals.TypedSignal

Signals.When

Signals.__init__

Signals._async_mode

Signals._debug_mode

Signals.action

Signals.activate_timer

Signals.add_child!

Signals.async_mode

Signals.async_signal — Function.

s = async_signal(f, args...; init = nothing)

Create a signal initialized to init whos action f(args...) will run asynchronously in a different task whenever its arguments update.async signals only work in a push based paradigm.

Signals.bind! — Function.

bind!(dest::Signal, src::Signal)

For every update to src also update dest with the same value.

Signals.buffer — Function.

s = buffer(input; buf_size = Inf, timespan = 1, type_stable = false)

Create a signal who buffers updates to signal input until maximum size of buf_size or until timespan seconds have passed. The signal value is the last full buffer emitted or an empty vector if the buffer have never been filled before.

Buffer type will be Any unless type_stable is set to true, then it will be set to the value of the first encountered item.

Signals.call_no_inline

Signals.clean_stacktrace

Base.count — Function.

count_signal(s::Signal)

Create a Signal that counts updates to input Signal s

Signals.debounce — Function.

debounce(f, args...; delay = 1, v0 = nothing)

Create a Signal whos action f(args...) will be called only after delay seconds have passed since the last time its args were updated. Only works in push based paradigm. If v0 is not specified then the initial value is f(args...).

Signals.debug_mode

Signals.droprepeats — Function.

droprepeats(input)

Drop updates to input whenever the new value is the same as the previous value of the signal.

Signals.echo — Function.

echo(s::Signal,name = "")

echos the value of signal s on each update, you can specify a name to distinguish between different echos

Signals.empty_queues

Signals.eval

Signals.eventloop

Signals.eventloop_cond

Signals.every — Function.

s = every(dt; duration = Inf)

A signal that updates every dt seconds to the current timestamp, for duration seconds.

Signals.foldp — Function.

fold over past values

foldp(op, v0 ,sig)

reduce the given signal sig with the given binary operator op. the value of the signal just after creation is op(v0,sig[])

Signals.for_signal — Function.

 s = for_signal(f::Function, range, args...; fps = 1)

Create a Signal that updates to f(i,args....) for i in range every 1/fps seconds. range and args can be of type Signal or any other type. The loop starts whenever one of the arguments updates. If the previous for loop did not complete it gets cancelled.

Example

rng = Signal(1:5)
A = Signal(2)
B = for_signal(rng,A;fps = 30) do i,a
    println(a^i)
end;

Signals.fps — Function.

s = fps(freq; duration = Inf)

A signal that updates freq times a second to the current timestamp, for duration seconds.

Signals.fpswhen — Function.

s = fpswhen(switch::Signal, freq; duration = Inf)

A signal that updates 'freq' times a second to the current timestamp, for duration seconds if and only if the value of switch is true.

Signals.handle_err

Signals.include

Signals.invalidate!

Signals.previous — Function.

previous(s::Signal)

Create a Signal that holds previous input to s

Signals.propagated

Signals.propogate!

Signals.pull!

Signals.pull_args

Signals.pull_enqueue

Signals.pull_queue

Signals.push_queue

Signals.recursion_free — Function.

recursion_free(f::Function,args...)

creates a Signal whos action f(args...) is protected against infinite loop recursion.

julia> A = Signal(1)
julia> B = recursion_free(A) do a
            A(a+1)
       end

julia> A(10)
10
julia> A[]
11

In the example above ,if recursion_free were to be subsituted with regular Signal it would result in an infinite loop in the non-async mode recursion_free protects against that

Signals.remote_signal — Function.

s = remote_signal(f, args...; init = nothing, procid = first(workers()))

Create a signal initialized to init whos action f(args...) will run remotely in a process with id procid, whenever its arguments update.remote signals only work in a push based paradigm.

Signals.run_till_now

Signals.sampleon — Function.

sampleon(A, B)

Sample the value of B whenever A updates.

Signals.set_value!

Signals.soft_push!

Signals.state — Function.

Retrieve the internal state of a Signal

Signals.store!

Signals.strict_push! — Function.

strict_push!(s::Signal, val, async = Signals.async_mode())

Set s to val and propagate into derived signals, is async is true(default) then updates to derived signals will occur asynchronically.

Signals.throttle — Function.

throttle(f::Function, args...; maxfps = 0.03)

Create a throttled Signal whos action f(args...) will be called only if 1/maxfps time has passed since the last time it updated. The resulting Signal will be updated maximum of maxfps times per second.

Signals.unbind! — Function.

unbind!(dst::Signal,src::Signal)

Remove bindings from src to dst that were previously created using bind.

unbind!(dst::Signal)

Remove all bindings that were previously created using bind that will cause dst to update.

Signals.valid

Signals.valid_args

Signals.validate

Signals.value

Signals.when — Function.

when(f, condition::Signal, args...; v0 = nothing)

Create a Signal that will update to f(args...) when any of its input args updates only if condition has value true. If condition != true in the time of creation the signal will be initialized to value v0.

Examples

julia> A = Signal(1)
julia> condition = Signal(A) do a
           a<10
       end
julia> B = when(condition,A) do a
       println("$a is smaller than 10")
   end
julia> A(1)
1 is smaller than 10
1

julia> A(12)
12

Signals.world_age