Docstrings

TypedTables.FlexTable — Type.

FlexTable(name1 = array1, ...)

Create a column-storage-based FlexTable with column names name1, etc, from arrays array1, etc. The input arrays array1, etc, must share the same dimensionality and indices.

FlexTable itself is an AbstractArray whose elements are NamedTuples of the form (name1 = first(array1), ...), etc. Rows of the table are obtained via standard array indexing table[i], and columns via table.name.

FlexTable differs from Table in that the columns are mutable - you may add, remove, rename and replace entire columns of a FlexTable, but not a Table. However, Table can access and iterate rows in local scope with fast, fully type-inferred code while FlexTable will be more efficient with a higher-order interface.

TypedTables.GetProperty

TypedTables.Table — Type.

Table(name1 = array1, ...)

Create a column-storage-based Table with column names name1, etc, from arrays array1, etc. The input arrays array1, etc, must share the same dimensionality and indices.

Table itself is an AbstractArray whose elements are NamedTuples of the form (name1 = first(array1), ...), etc. Rows of the table are obtained via standard array indexing table[i], and columns via table.name.

Table differs from FlexTable in that the columns are immutable - you may add, remove, rename and replace entire columns of a FlexTable, but not a Table. However, Table can access and iterate rows in local scope with fast, fully type-inferred code while FlexTable will be more efficient with a higher-order interface.

TypedTables.TypedTables

TypedTables.__getindex_dims

TypedTables.__ndims

TypedTables._columns

TypedTables._eltypes

TypedTables._flextable

TypedTables._getindex_dims

TypedTables._hcat_ndims

TypedTables._ndims

TypedTables._similar

TypedTables._table

TypedTables._vcat_ndims

TypedTables.balance_widths

TypedTables.balance_widths!

TypedTables.columnnames — Function.

columnnames(table)

Return a tuple of the column names of a Table.

columnnames(table)

Return a tuple of the column names of a Table.

Tables.columns — Function.

columns(table::Table)

Convert a Table into a NamedTuple of its columns.

columns(dataframe::FlexTable)

Convert a FlexTable into a NamedTuple of its columns.

TypedTables.compact_string

TypedTables.compact_string_row

TypedTables.eval

TypedTables.include

TypedTables.removenothings

TypedTables.rows

TypedTables.showtable

TypedTables.truncate_string