Docstrings

SimpleWeightedGraphs.AbstractSimpleWeightedEdge

SimpleWeightedGraphs.AbstractSimpleWeightedGraph — Type.

AbstractSimpleWeightedGraph

An abstract type representing a simple graph structure. AbstractSimpleWeightedGraphs must have the following elements:

weightmx::AbstractSparseMatrix{Real}

SimpleWeightedGraphs.FIXEDSTR

SimpleWeightedGraphs.SWGFormat

SimpleWeightedGraphs.SWGHeader

SimpleWeightedGraphs.SimpleWeightedDiGraph — Type.

SimpleWeightedDiGraph{T, U}

A type representing a directed graph with weights of type U.

Note that adding or removing vertices or edges is not particularly performant; see MetaGraphs.jl for possible alternatives.

SimpleWeightedGraphs.SimpleWeightedDiGraphEdge

SimpleWeightedGraphs.SimpleWeightedEdge

SimpleWeightedGraphs.SimpleWeightedGraph — Type.

SimpleWeightedGraph{T, U}

A type representing an undirected graph with weights of type U.

Note that adding or removing vertices or edges is not particularly performant; see MetaGraphs.jl for possible alternatives.

SimpleWeightedGraphs.SimpleWeightedGraphEdge

SimpleWeightedGraphs.SimpleWeightedGraphs

SimpleWeightedGraphs.WDiGraph

SimpleWeightedGraphs.WGraph

SimpleWeightedGraphs._cc

SimpleWeightedGraphs._parse_header

SimpleWeightedGraphs._swg_read_one_graph

SimpleWeightedGraphs._swg_skip_one_graph

SimpleWeightedGraphs.eval

SimpleWeightedGraphs.get_weight

SimpleWeightedGraphs.include

SimpleWeightedGraphs.loadswg

SimpleWeightedGraphs.loadswg_mult — Function.

loadswg_mult(io)

Return a dictionary of (name=>graph) loaded from IO stream io.

SimpleWeightedGraphs.saveswg — Function.

saveswg(io, g, gname)

Write a graph g with name gname in a proprietary format to the IO stream designated by io. Return 1 (number of graphs written).

SimpleWeightedGraphs.saveswg_mult — Function.

saveswg_mult(io, graphs)

Write a dictionary of (name=>graph) to an IO stream io, with default GZip compression. Return number of graphs written.

SimpleWeightedGraphs.weight

SimpleWeightedGraphs.weighttype