Docstrings

QuickTypes.@qimmutable

QuickTypes.@qmutable — Macro.

Quick mutable struct definition. See ?@qstruct

QuickTypes.@qmutable_fp — Macro.

Fully-parametric version of @qmutable. @qmutable_fp Foo(a, b=2) is like @qmutable Foo{T, U}(a::T, B::U=2)

QuickTypes.@qmutable_np

QuickTypes.@qstruct — Macro.

Quick type definition.

@qstruct Car(size, nwheels::Int=4; brand::String="unnamed") <: Vehicle

expands into

struct Car <: Vehicle
    size
    nwheels::Int
    brand::String
    Car(size, nwheels=4; brand="unnamed") = new(size, nwheels, brand)
end

Also supports parametric types: @qstruct Door{T}(size::T). Invariants can be enforced using do-syntax:

@qstruct Human(name, height::Float64) do
    @assert height > 0    # arbitrary code, executed before creating the object
end

Note: @qstruct automatically defines a Base.show method for the new type, unless _define_show=false (eg. @qstruct(x, y; _define_show=false)).

QuickTypes.@qstruct_fp — Macro.

Fully-parametric version of @qstruct. @qstruct_fp Foo(a, b=2) is like @qstruct Foo{T, U}(a::T, B::U=2)

QuickTypes.@qstruct_np — Macro.

Narrowly-parametric version of @qstruct. @qstruct_np Foo(a, b=2) is like @qstruct Foo{T, U}(a::T, B::U=2), but it will additionally specialize on types: Foo(Int, 2.0) ==> Foo{Type{Int64},Float64}(Int64, 2.0)

QuickTypes.@qtype

QuickTypes.QuickTypes

QuickTypes.all_type_vars_present

QuickTypes.build_show_def — Function.

Build the Base.show function definition for that type

QuickTypes.construct — Function.

For a type X defined with @qmutable/@qstruct and with fields a, b, c, ..., QuickTypes.construct(X, 1, 2, 3...) is a purely-positional constructor of X. This is useful for writing generic structure traversal. For any object of a type defined by @qstruct, this holds:

QuickTypes.construct(QuickTypes.roottypeof(o), QuickTypes.fieldsof(o)...) == o

QuickTypes.eval

QuickTypes.fieldsof — Function.

fieldsof(obj) returns the fields of obj in a tuple. See also QuickTypes.construct

QuickTypes.fieldtypes — Function.

fieldtypes(typ) returns the types of the fields of a composite type.

QuickTypes.get_sym

QuickTypes.include

QuickTypes.is_mutable

QuickTypes.make_parametric

QuickTypes.narrow_typeof

QuickTypes.parse_funcall — Function.

parse_funcall(fcall)

Parses fname(args; kwargs). Returns (fname, args, kwargs, constraints)

QuickTypes.qexpansion

QuickTypes.roottype — Function.

roottype(typ::Type) returns the parameterless type. Eg. roottype(X{A}) => X

QuickTypes.roottypeof — Function.

roottypeof(obj) returns the type of obj with generic parametric types. Eg. roottypeof(a::SomeType{Int}) -> SomeType{T}. See QuickTypes.construct

QuickTypes.special_kwargs

QuickTypes.tuple_parameters — Function.

tuple_parameters{T<:Tuple}(::Type{T}) returns the type of each element of the tuple, as svec(type1, type2, ...)

QuickTypes.tuple_parameters_stable

QuickTypes.type_length

QuickTypes.type_parameters

QuickTypes.type_simple_name