Docstrings

UnicodePlots.AsciiCanvas — Type.

As the name suggests the AsciiCanvas only uses ASCII characters to draw it's content. Naturally, it doesn't look quite as nice as the Unicode-based ones. However, in some situations it might yield better results. Printing plots to a file is one of those situations.

The AsciiCanvas is best utilized in combination with lineplot. For scatterplot we suggest to use the DotCanvas instead.

UnicodePlots.BarplotGraphics

UnicodePlots.BlockCanvas — Type.

The BlockCanvas is also Unicode-based. It has half the resolution of the BrailleCanvas. In contrast to BrailleCanvas, the pixels don't have visible spacing between them. This canvas effectively turns every character into 4 pixels that can individually be manipulated using binary operations.

UnicodePlots.BoxplotGraphics

UnicodePlots.BrailleCanvas — Type.

The type of canvas with the highest resolution for Unicode-based plotting. It uses the Unicode characters for the Braille symbols to represent individual pixel. This effectively turns every character into 8 pixels that can individually be manipulated using binary operations.

UnicodePlots.Canvas

UnicodePlots.CreateLookupCanvas

UnicodePlots.DOC_PLOT_PARAMS

UnicodePlots.DensityCanvas — Type.

Unlike the BrailleCanvas, the density canvas does not simply mark a "pixel" as set. Instead it increments a counter per character that keeps track of the frequency of pixels drawn in that character. Together with a variable that keeps track of the maximum frequency, the canvas can thus draw the density of datapoints.

UnicodePlots.DotCanvas — Type.

Similar to the AsciiCanvas, the DotCanvas only uses ASCII characters to draw it's content. Naturally, it doesn't look quite as nice as the Unicode-based ones. However, in some situations it might yield better results. Printing plots to a file is one of those situations.

The DotCanvas is best utilized in combination with scatterplot. For lineplot we suggest to use the AsciiCanvas instead.

UnicodePlots.FiveNumberSummary

UnicodePlots.GraphicsArea

UnicodePlots.LookupCanvas

UnicodePlots.Plot — Type.

Plot(graphics; nargs...)

Description

Decoration for objects that are GraphicsArea (or Canvas). It is used to surround the inner GraphicsArea object with additional information such as a title, border, and axis labels.

Usage

Plot(graphics; title = "", xlabel = "", ylabel = "", border = :solid, margin = 3, padding = 1, labels = true)

Plot(x, y, canvastype; title = "", xlabel = "", ylabel = "", width = 40, height = 15, border = :solid, xlim = (0, 0), ylim = (0, 0), margin = 3, padding = 1, labels = true, grid = true)

Arguments

graphics : The GraphicsArea (e.g. a subtype of Canvas) that the plot should decorate.

x : The horizontal position for each point.

y : The vertical position for each point.

canvastype : The type of canvas that should be used for drawing.

title : Text to display on the top of the plot.

xlabel : Text to display on the x axis of the plot

ylabel : Text to display on the y axis of the plot

labels : Boolean. Can be used to hide the labels by setting labels = false.

border : The style of the bounding box of the plot. Supports :corners, :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :green, :blue, :red, :yellow, :cyan, :magenta, :white, :normal

width : Number of characters per row that should be used for plotting.

height : Number of character rows that should be used for plotting.

xlim : Plotting range for the x axis. [0, 0] stands for automatic.

ylim : Plotting range for the y axis. [0, 0] stands for automatic.

grid : If true, draws grid-lines at the origin.

Methods

title!(plot::Plot, title::String)

xlabel!(plot::Plot, xlabel::String)

ylabel!(plot::Plot, xlabel::String)

annotate!(plot::Plot, where::Symbol, value::String)

annotate!(plot::Plot, where::Symbol, row::Int, value::String)

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

see also

scatterplot, lineplot, BarplotGraphics, BrailleCanvas, BlockCanvas, AsciiCanvas

UnicodePlots.UnicodePlots

UnicodePlots._findnz

UnicodePlots._lineplot

UnicodePlots._nocolor_string

UnicodePlots.addrow!

UnicodePlots.addseries!

UnicodePlots.annotate! — Function.

annotate!(plot, where, value, [color])

annotate!(plot, where, row, value, [color])

This method is responsible for the setting all the textual decorations of a plot.

Note that where can be any of: :tl (top-left), :t (top-center), :tr (top-right), :bl (bottom-left), :b (bottom-center), :br (bottom-right), :l (left), :r (right).

If where is either :l, or :r, then row can be between 1 and the number of character rows of the plots canvas.

UnicodePlots.ascii_decode

UnicodePlots.ascii_lookup

UnicodePlots.ascii_signs

UnicodePlots.barplot — Function.

barplot(text, heights; nargs...)

Description

Draws a horizontal barplot. It uses the first parameter (text) to denote the names for the bars, and the second parameter (heights) as their values. This means that the two vectors have to have the same length. Alternatively, one can specify a barplot using a dictionary dict. In that case, the keys will be used as the names and the values, which have to be numeric, will be used as the heights of the bars.

Usage

barplot(text, heights; xscale = identity, title = "", xlabel = "", ylabel = "", labels = true, border = :barplot, margin = 3, padding = 1, color = :green, width = 40, symb = "■")

barplot(dict; kwargs...)

Arguments

text : The labels / captions of the bars.

heights : The values / heights of the bars.

dict : A dictonary in which the keys will be used as text and the values will be utilized as heights.

xscale : Function to transform the bar length before plotting. This effectively scales the x-axis without influencing the captions of the individual bars. e.g. use xscale = log10 for logscale.

title : Text to display on the top of the plot.

xlabel : Text to display on the x axis of the plot

ylabel : Text to display on the y axis of the plot

labels : Boolean. Can be used to hide the labels by setting labels = false.

border : The style of the bounding box of the plot. Supports :corners, :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :green, :blue, :red, :yellow, :cyan, :magenta, :white, :normal

width : Number of characters per row that should be used for plotting.

symb : Specifies the character that should be used to render the bars

Returns

A plot object of type Plot{BarplotGraphics}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Examples

julia> barplot(["Paris", "New York", "Moskau", "Madrid"],
               [2.244, 8.406, 11.92, 3.165],
               xlabel = "population [in mil]")
            ┌                                        ┐
      Paris ┤■■■■■■ 2.244
   New York ┤■■■■■■■■■■■■■■■■■■■■■■■ 8.406
     Moskau ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 11.92
     Madrid ┤■■■■■■■■■ 3.165
            └                                        ┘
                       population [in mil]

See also

Plot, histogram, BarplotGraphics

UnicodePlots.barplot! — Function.

barplot!(plot, text, heights) -> plot

Mutating variant of barplot, in which the first parameter (plot) specifies the existing plot to draw additional bars on.

See barplot for more information.

UnicodePlots.block_decode

UnicodePlots.block_signs

UnicodePlots.border_ascii

UnicodePlots.border_barplot

UnicodePlots.border_bold

UnicodePlots.border_corners

UnicodePlots.border_dashed

UnicodePlots.border_dotted

UnicodePlots.border_none

UnicodePlots.border_solid

UnicodePlots.bordermap

UnicodePlots.boxplot — Function.

boxplot(data; kwargs...)

Description

Draws a box-and-whisker plot.

The first argument specifies the data to plot. This is a vector of vectors, with each inner vector representing a data series. We use a vector of vectors over a matrix to allow series of different lengths. Optionally, a list of text may be provided, with length equal to the number of series.

Alternatively, one can specify a boxplot using a dictionary. In that case, the values, which have to be numeric, will be used as the data series, and the keys, which have to be strings, will be used as the labels.

Usage

boxplot([text], data; title = "", xlabel = "", ylabel = "", labels = true, border = :corners, margin = 3, padding = 1, color = :green, width = 40, xlim)

boxplot(dictionary; nargs...)

Arguments

text : Optional. The labels/captions of the boxes.

data : The data the box plot is based on. A vector of vectors, with each inner vector representing a data series. Choose a vector of vectors over a matrix to allow series of different lengths.

dictionary : A dictonary in which the keys will be used as text and the values will be utilized as data.

title : Text to display on the top of the plot.

xlabel : Text to display on the x axis of the plot

ylabel : Text to display on the y axis of the plot

labels : Boolean. Can be used to hide the labels by setting labels = false.

border : The style of the bounding box of the plot. Supports :corners, :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :green, :blue, :red, :yellow, :cyan, :magenta, :white, :normal

width : Number of characters per row that should be used for plotting.

xlim : Plotting range for the data axis. (0, 0) stands for automatic.

Returns

A plot object of type Plot{BoxplotGraphics}

Author(s)

Matthew Lake (Github: https://github.com/mgtlake)

Examples

julia> boxplot([1, 2, 3, 7], title = "Test")
                      Test
    ┌                                        ┐
     ╷   ┌────┬─────────┐                   ╷
     ├───┤    │         ├───────────────────┤
     ╵   └────┴─────────┘                   ╵
    └                                        ┘
    1                   4                    7

See also

Plot, histogram, BoxplotGraphics

UnicodePlots.boxplot! — Function.

boxplot!(plot, data; nargs) → Plot

Mutating variant of boxplot, in which the first parameter (plot) specifies the existing plot to draw on.

See boxplot for more information.

UnicodePlots.braille_signs

UnicodePlots.ceil_neg_log10

UnicodePlots.color_cycle

UnicodePlots.color_decode

UnicodePlots.color_encode

UnicodePlots.colors

UnicodePlots.compute_stair_lines

UnicodePlots.den_signs

UnicodePlots.densityplot — Function.

densityplot(x, y; kwargs...)

Description

Draws a density plot for the given points.

The first vector x should contain the horizontal positions for all the points. The second vector y should then contain the corresponding vertical positions respectively. This means that the two vectors must be of the same length and ordering.

Usage

densityplot(x, y; name = "", title = "", xlabel = "", ylabel = "", labels = true, border = :solid, margin = 3, padding = 1, color = :auto, width = 40, height = 15, xlim = (0, 0), ylim = (0, 0), grid = false)

Arguments

x : The horizontal position for each point. If omitted, the axes of y will be used as x.

y : The vertical position for each point.

name : Annotation of the current drawing to be displayed on the right

title : Text to display on the top of the plot.

xlabel : Text to display on the x axis of the plot

ylabel : Text to display on the y axis of the plot

labels : Boolean. Can be used to hide the labels by setting labels = false.

border : The style of the bounding box of the plot. Supports :corners, :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :green, :blue, :red, :yellow, :cyan, :magenta, :white, :normal

width : Number of characters per row that should be used for plotting.

height : Number of character rows that should be used for plotting.

xlim : Plotting range for the x axis. (0, 0) stands for automatic.

ylim : Plotting range for the y axis. (0, 0) stands for automatic.

grid : If true, draws grid-lines at the origin.

Returns

A plot object of type Plot{DensityCanvas}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Examples

julia> densityplot(randn(1000), randn(1000), title = "Density Plot")
                      Density Plot
        ┌────────────────────────────────────────┐
    2.9 │                    ░                   │
        │                ░    ░  ░               │
        │                ░░░  ░ ░ ░      ░       │
        │             ░░░ ░░▒░▒░░░ ▒ ░           │
        │          ░░ ░░ ░▒░░▒▒▒▒░░░░░░░         │
        │         ░░ ░▒░░░░░▓▒▒░▒░▒░▓░▒░░░       │
        │        ░ ░░░░▓▓▒▓▒▒█▒▓▒▒▓▒▒░▒ ░ ░      │
        │           ░ ▒▒▒▒▓▓▒▓▓▓▓▒█▒▒░░░░░░ ░    │
        │     ░   ░░░░░░░▒▒▓░░▓▒█▒▒▓▓▒▒ ░    ░   │
        │           ░ ░░▒▒░░▒░░▒░░░░ ░░  ░ ░     │
        │         ░   ░ ░ ░░░░░░ ░░ ░░ ░         │
        │              ░  ░ ░ ░  ░ ░   ░░░       │
        │                ░░  ░                   │
        │                                        │
   -3.3 │                                        │
        └────────────────────────────────────────┘
        -3.4                                   2.9

See also

Plot, scatterplot, DensityCanvas

UnicodePlots.densityplot!

UnicodePlots.dot_decode

UnicodePlots.dot_signs

UnicodePlots.eval

UnicodePlots.extend_limits

UnicodePlots.float_round_log10

UnicodePlots.grid

UnicodePlots.height

UnicodePlots.histogram — Function.

histogram(data; kwargs...)

Description

Draws a horizontal histogram of the given data. The positional parameter data can either be a StatsBase.Histogram, or some AbstractArray. In the later case the Histogram will be fitted automatically.

Note internally that histogram is a simply wrapper for barplot, which means that it supports the same keyword arguments.

Usage

histogram(x; nbins, closed = :left, kwargs...)

histogram(hist; xscale = identity, title = "", xlabel = "", ylabel = "", labels = true, border = :barplot, margin = 3, padding = 1, color = :green, width = 40, symb = "▇")

Arguments

x : Array of numbers for which the histogram should be computed.

nbins : The approximate number of bins that should be used.

closed : If :left (the default), the bin intervals are left-closed $[a,b)$; if :right, intervals are right-closed $(a,b]$.

hist : A fitted StatsBase.Histogram that should be plotted.

xscale : Function to transform the bar length before plotting. This effectively scales the x-axis without influencing the captions of the individual bars. e.g. use xscale = log10 for logscale.

title : Text to display on the top of the plot.

xlabel : Text to display on the x axis of the plot

ylabel : Text to display on the y axis of the plot

labels : Boolean. Can be used to hide the labels by setting labels = false.

border : The style of the bounding box of the plot. Supports :corners, :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :green, :blue, :red, :yellow, :cyan, :magenta, :white, :normal

width : Number of characters per row that should be used for plotting.

symb : Specifies the character that should be used to render the individual bars.

Returns

A plot object of type Plot{BarplotGraphics}

Author(s)

Iain Dunning (Github: https://github.com/IainNZ)

Christof Stocker (Github: https://github.com/Evizero)

Kenta Sato (Github: https://github.com/bicycle1885)

Examples

julia> histogram(randn(1000) * 0.1, closed = :right, nbins = 15)
                  ┌                                        ┐
   (-0.3 , -0.25] ┤▇ 4
   (-0.25, -0.2 ] ┤▇▇ 12
   (-0.2 , -0.15] ┤▇▇▇▇▇▇▇▇▇ 48
   (-0.15, -0.1 ] ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 114
   (-0.1 , -0.05] ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 143
   (-0.05, -0.0 ] ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 177
   ( 0.0 ,  0.05] ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 192
   ( 0.05,  0.1 ] ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 152
   ( 0.1 ,  0.15] ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 83
   ( 0.15,  0.2 ] ┤▇▇▇▇▇▇▇▇ 45
   ( 0.2 ,  0.25] ┤▇▇▇▇ 23
   ( 0.25,  0.3 ] ┤▇ 6
   ( 0.3 ,  0.35] ┤ 1
                  └                                        ┘
                                  Frequency

See also

Plot, barplot, BarplotGraphics

UnicodePlots.include

UnicodePlots.lineplot — Function.

lineplot(x, y; kwargs...)

Description

Draws a path through the given points on a new canvas.

The first (optional) vector x should contain the horizontal positions for all the points along the path. The second vector y should then contain the corresponding vertical positions respectively. This means that the two vectors must be of the same length and ordering.

Usage

lineplot([x], y; name = "", title = "", xlabel = "", ylabel = "", labels = true, border = :solid, margin = 3, padding = 1, color = :auto, width = 40, height = 15, xlim = (0, 0), ylim = (0, 0), canvas = BrailleCanvas, grid = true)

lineplot(fun, [start], [stop]; kwargs...)

Arguments

x : Optional. The horizontal position for each point. Can be a real number or of type TimeType. If omitted, the axes of y will be used as x.

y : The vertical position for each point.

fun : A unary function $f: R -> R$ that should be evaluated, and drawn as a path from start to stop (numbers in the domain).

name : Annotation of the current drawing to be displayed on the right.

title : Text to display on the top of the plot.

xlabel : Text to display on the x axis of the plot

ylabel : Text to display on the y axis of the plot

labels : Boolean. Can be used to hide the labels by setting labels = false.

border : The style of the bounding box of the plot. Supports :corners, :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :green, :blue, :red, :yellow, :cyan, :magenta, :white, :normal

width : Number of characters per row that should be used for plotting.

height : Number of character rows that should be used for plotting.

xlim : Plotting range for the x axis. (0, 0) stands for automatic.

ylim : Plotting range for the y axis. (0, 0) stands for automatic.

canvas : The type of canvas that should be used for drawing.

grid : If true, draws grid-lines at the origin.

Returns

A plot object of type Plot{T<:Canvas}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Milktrader (Github: https://github.com/milktrader)

Examples

julia> lineplot([1, 2, 7], [9, -6, 8], title = "My Lineplot")
                      My Lineplot
       ┌────────────────────────────────────────┐
    10 │⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⢇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡠│
       │⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠤⠊⠁⠀│
       │⠀⢣⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⠔⠊⠁⠀⠀⠀⠀│
       │⠀⠈⡆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⠔⠊⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡠⠔⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⢇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡠⠒⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠤⠤⠤⠼⡤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⢤⠤⠶⠥⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤│
       │⠀⠀⠀⠀⢣⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠤⠊⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠈⡆⠀⠀⠀⠀⠀⠀⠀⣀⠔⠊⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⢱⠀⠀⠀⠀⡠⠔⠊⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⢇⡠⠔⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠈⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
   -10 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       └────────────────────────────────────────┘
       1                                        7

See also

Plot, scatterplot, stairs, BrailleCanvas, BlockCanvas, AsciiCanvas, DotCanvas

UnicodePlots.lineplot!

UnicodePlots.lines!

UnicodePlots.lookup_decode

UnicodePlots.lookup_encode

UnicodePlots.ncols

UnicodePlots.next_color!

UnicodePlots.nrows

UnicodePlots.origin

UnicodePlots.origin_x

UnicodePlots.origin_y

UnicodePlots.pixel!

UnicodePlots.pixel_height

UnicodePlots.pixel_size

UnicodePlots.pixel_width

UnicodePlots.plotting_range

UnicodePlots.plotting_range_narrow

UnicodePlots.points!

UnicodePlots.print_border_bottom

UnicodePlots.print_border_top

UnicodePlots.print_color

UnicodePlots.print_title

UnicodePlots.printrow

UnicodePlots.round_down_subtick

UnicodePlots.round_down_tick

UnicodePlots.round_neg_log10

UnicodePlots.round_up_subtick

UnicodePlots.round_up_tick

UnicodePlots.roundable

UnicodePlots.scatterplot — Function.

scatterplot(x, y; kwargs...)

Description

Draws the given points on a new canvas.

The first (optional) vector x should contain the horizontal positions for all the points. The second vector y should then contain the corresponding vertical positions respectively. This means that the two vectors must be of the same length and ordering.

Usage

scatterplot([x], y; name = "", title = "", xlabel = "", ylabel = "", labels = true, border = :solid, margin = 3, padding = 1, color = :auto, width = 40, height = 15, xlim = (0, 0), ylim = (0, 0), canvas = BrailleCanvas, grid = true)

Arguments

x : Optional. The horizontal position for each point. If omitted, the axes of y will be used as x.

y : The vertical position for each point.

name : Annotation of the current drawing to be displayed on the right

title : Text to display on the top of the plot.

xlabel : Text to display on the x axis of the plot

ylabel : Text to display on the y axis of the plot

labels : Boolean. Can be used to hide the labels by setting labels = false.

border : The style of the bounding box of the plot. Supports :corners, :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :green, :blue, :red, :yellow, :cyan, :magenta, :white, :normal

width : Number of characters per row that should be used for plotting.

height : Number of character rows that should be used for plotting.

xlim : Plotting range for the x axis. (0, 0) stands for automatic.

ylim : Plotting range for the y axis. (0, 0) stands for automatic.

canvas : The type of canvas that should be used for drawing.

grid : If true, draws grid-lines at the origin.

Returns

A plot object of type Plot{T<:Canvas}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Examples

julia> scatterplot(randn(50), randn(50), title = "My Scatterplot")
                   My Scatterplot
      ┌────────────────────────────────────────┐
    3 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡆⠀⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⡇⠀⠀⠀⠀⠀⠀⠠⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠡⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⡀⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠀⠀⠁⠀⡇⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠀⠂⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠐⠄⠀⠀⠀⠀⡀⠀⠂⡇⠀⠀⠀⠂⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠤⠤⠤⠤⠤⠤⠴⠤⠤⠤⠬⠤⠬⠤⠤⠤⡧⢥⡤⠤⠤⡤⠤⠤⠤⠬⠤⠤⠤⠤⠤⠤⠤⠤⡤⠤⠤⠤⠤⠄│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠁⠀⠀⠀⠀⠀⠀⠀⡗⠐⠨⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⡇⠀⠐⠀⠨⠀⠀⠂⠀⠀⠀⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠠⠀⠄⠀⠀⠀⠀⠀⡇⠀⠀⠀⠐⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠁⠀⠀⠀⠀⠀⠀⠀⠀⠁⠂⠰⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
   -3 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      └────────────────────────────────────────┘
      -2                                       3

See also

Plot, lineplot, stairs, BrailleCanvas, BlockCanvas, AsciiCanvas, DotCanvas

UnicodePlots.scatterplot!

UnicodePlots.spy — Function.

spy(A; kwargs...)

Description

Plots the sparsity pattern for the given matrix A. This means that a scatterplot that resembles the matrix is drawn, in which only the pixel for non-zero elements of the matrix are set.

If the parameters width and height are not explicitly specified, then the function will attempt to preserve the aspect ratio of the matrix, while also attempting to fit the resulting plot withing the bounding box specified by maxwidth and maxheight

Usage

spy(A; maxwidth = 70, maxheight = 40, title = "Sparsity Pattern", labels = true, border = :solid, margin = 3, padding = 1, color = :auto, width = 0, height = 0, canvas = BrailleCanvas, grid = true)

Arguments

A : The matrix of interest for which non-zero elements should be drawn

maxwidth : Maximum number of characters per row that should be used for plotting

maxheight : Maximum number of character rows that should be used for plotting.

title : Text to display on the top of the plot.

labels : Boolean. Can be used to hide the labels by setting labels = false.

border : The style of the bounding box of the plot. Supports :corners, :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :green, :blue, :red, :yellow, :cyan, :magenta, :white, :normal.

width : Exact number of characters per row that should be used for plotting. 0 stands for automatic.

height : Exact number of character rows that should be used for plotting. 0 stands for automatic.

canvas : The type of canvas that should be used for drawing.

Returns

A plot object of type Plot{T<:Canvas}

Author(s)

Dominique (Github: https://github.com/dpo)

Christof Stocker (Github: https://github.com/Evizero)

Jake Bolewski (Github: https://github.com/jakebolewski)

Examples

julia> spy(sprandn(50, 120, .05))
                           Sparsity Pattern
      ┌────────────────────────────────────────────────────────────┐
    1 │⢀⠀⠒⠀⢄⠂⠀⡀⠀⠀⠀⠰⠀⠐⠀⠀⠀⠀⠂⠀⠀⠀⠠⠰⠀⠀⠀⡀⠄⠀⠠⠀⠀⠀⠀⠀⡀⠀⠀⠀⡀⠐⠀⠐⠀⠀⡀⠀⠀⠀⠀⠀⠀⠂⡀⠀⡀⠠⠀⠀│ > 0
      │⠀⠀⠀⠀⠀⢀⢀⠀⠀⠀⠀⠐⠄⠀⢀⠂⠀⠠⠀⠉⠀⠀⠀⠀⢐⠀⠀⠀⠀⠀⠠⠐⠄⠄⠀⠀⠀⠐⠐⠈⢁⠀⠀⠀⠀⠀⠀⠀⠠⠀⠐⡐⠀⠄⠀⠀⠈⠀⠀⠠│ < 0
      │⡀⠀⠀⠀⡀⠊⠀⠀⠀⢀⢀⠀⠄⠀⠂⡠⠂⠐⠐⠀⠀⠄⠀⠀⠀⠄⠀⠀⠀⠀⡀⠀⠀⠁⠀⠀⠀⠀⠀⠀⠠⠀⠂⠀⠠⠀⠀⠀⠆⠄⠈⠀⢀⠀⠀⠀⠐⠄⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⢠⠀⢀⠀⠀⠀⢀⠀⠄⠀⠀⠀⠂⠀⠀⠀⠇⠀⢀⠀⠀⠁⠁⠀⠄⠄⠀⠀⠁⠁⠀⡈⠀⠁⠄⠀⠀⠄⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠀│
      │⠀⠀⠀⠀⠀⠀⠀⠀⠀⡐⢀⠀⠀⠀⠐⠂⠁⢀⠠⠀⠀⠀⠀⠀⠁⠀⠀⠀⠠⠀⠀⠀⠀⢀⠀⢄⠀⠀⠀⠐⠀⠀⠈⠀⠀⠐⠀⠄⠂⠠⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠠⠀⠠⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠠⠈⠀⠀⠀⠀⢀⠀⠀⠄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
      │⠐⠀⠉⠐⠀⠀⠁⠀⠀⠀⠀⠀⠀⠌⠀⠀⠀⠄⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠌⠀⠀⠄⠀⠀⡀⠀⠀⠀⠀⠀⢀⠈⠄⠀⡀⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀│
      │⠂⠀⠀⠀⠀⠀⠀⠠⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠀⡨⠀⠠⠀⠀⠀⠂⠀⠀⠀⠀⠀⠀⠈⡠⠀⡀⠀⠀⠁⠀⠁⠠⠀⠀⠀⠀⠀⠀⠀⠄⡀⠀⠄⡠⠀⠀⠀⠀│
      │⠐⠀⠀⠀⠀⠀⠀⠃⠀⠀⠀⠀⠁⠀⡀⠁⠀⢂⠂⠀⠀⠀⠈⠁⠀⠀⠁⠁⠄⠀⠀⠀⠀⠀⡀⠀⠀⠀⠂⠀⠀⠀⠀⠐⠀⠂⠀⠂⠁⠀⠆⠈⠀⠀⠄⠄⠀⠀⠀⠀│
      │⠀⠀⠀⠠⠀⠁⠀⠀⠀⠀⠂⠀⠠⠀⠀⠌⠀⠀⠂⠄⠀⡀⠂⠀⠀⡀⠠⠁⠄⠀⠀⠀⠀⠀⠄⠈⠀⢂⠐⠀⢀⠀⠠⠀⠄⠁⡀⠀⠀⠐⠀⠀⡐⠀⠀⠀⠁⠀⠀⠔│
      │⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠁⠃⡀⡁⠀⠀⠀⢀⠀⠀⠂⠀⠀⠠⠀⠀⠠⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠀⠬⠀⡀│
      │⠐⠀⠀⠀⠀⠀⠀⠀⠢⠀⠀⠀⠀⠀⢀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠀⠈⠁⠁⠀⠀⠄⠀⠀⠐⠀⠄⠀⠠⠀⠁⡀⠁⠀⠀⠠⠡⠀⠀⠀⠀⠀⠀⠀⠐⠄⠀⠀⠀⠀⡀│
   50 │⠀⠀⠀⠀⠁⠀⠀⠠⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠠⠂⠀⠀⠀⠀⠀⠀⠐⠀⠈⠂⠀⠀⠀⠀⠀⠀⠀⠀⠠⠈⠀⠀⠀⠀⠀⠀⠂⠀⠄⠀⠀⠀⠂⠁⠀⠀⠀⠂⠀│
      └────────────────────────────────────────────────────────────┘
      1                                                          120
                               nz = 275

See also

Plot, scatterplot, BrailleCanvas, BlockCanvas, AsciiCanvas, DotCanvas

UnicodePlots.stairs — Function.

stairs(x, y; kwargs...)

Description

Draws a staircase plot on a new canvas.

The first (optional) vector x should contain the horizontal positions for all the points. The second vector y should then contain the corresponding vertical positions respectively. This means that the two vectors must be of the same length and ordering.

Usage

stairs(x, y; style = :post, name = "", title = "", xlabel = "", ylabel = "", labels = true, border = :solid, margin = 3, padding = 1, color = :auto, width = 40, height = 15, xlim = (0, 0), ylim = (0, 0), canvas = BrailleCanvas, grid = true)

Arguments

x : The horizontal position for each point.

y : The vertical position for each point.

style : Specifies where the transition of the stair takes plays. Can be either :pre or :post.

name : Annotation of the current drawing to be displayed on the right.

title : Text to display on the top of the plot.

xlabel : Text to display on the x axis of the plot

ylabel : Text to display on the y axis of the plot

labels : Boolean. Can be used to hide the labels by setting labels = false.

border : The style of the bounding box of the plot. Supports :corners, :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :green, :blue, :red, :yellow, :cyan, :magenta, :white, :normal

width : Number of characters per row that should be used for plotting.

height : Number of character rows that should be used for plotting.

xlim : Plotting range for the x axis. (0, 0) stands for automatic.

ylim : Plotting range for the y axis. (0, 0) stands for automatic.

canvas : The type of canvas that should be used for drawing.

grid : If true, draws grid-lines at the origin.

Returns

A plot object of type Plot{T<:Canvas}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Dominique (Github: https://github.com/dpo)

Examples

julia> stairs([1, 2, 4, 7, 8], [1, 3, 4, 2, 7], style = :post, title = "My Staircase Plot")
                 My Staircase Plot
     ┌────────────────────────────────────────┐
   7 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀ ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⡄⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⢸⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢸│
     │⠀⠀⠀⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠧⠤⠤⠤⠤⠼│
     │⠀⠀⠀⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
   1 │⣀⣀⣀⣀⣀⣸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     └────────────────────────────────────────┘
     1                                        8

See also

Plot, scatterplot, lineplot, BrailleCanvas, BlockCanvas, AsciiCanvas, DotCanvas

UnicodePlots.stairs!

UnicodePlots.title — Function.

title(plot) -> String

Returns the current title of the given plot. Alternatively, the title can be changed with title!.

UnicodePlots.title! — Function.

title!(plot, newtitle) -> plot

Sets a new title for the given plot. Alternatively, the current title can be queried using title.

UnicodePlots.transform_name

UnicodePlots.width

UnicodePlots.x_pixel_per_char

UnicodePlots.xlabel — Function.

xlabel(plot) -> String

Returns the current label for the x-axis. Alternatively, the x-label can be changed with xlabel!

UnicodePlots.xlabel! — Function.

xlabel!(plot, newlabel) -> plot

Sets a new x-label for the given plot. Alternatively, the current label can be queried using xlabel

UnicodePlots.y_pixel_per_char

UnicodePlots.ylabel — Function.

ylabel(plot) -> String

Returns the current label for the y-axis. Alternatively, the y-label can be changed with ylabel!

UnicodePlots.ylabel! — Function.

ylabel!(plot, newlabel) -> plot

Sets a new y-label for the given plot. Alternatively, the current label can be queried using ylabel