Docstrings

SoftGlobalScope.@softscope — Macro.

@softscope(expr)

Apply "soft" scoping rules to the argument of the macro. For example

julia> s = 0;

julia> @softscope for i = 1:10
           s += i
       end

julia> s
55

SoftGlobalScope.SoftGlobalScope — Module.

SoftGlobalScope is a package that simplifies the variable scoping rules for code in global scope. It is intended for interactive shells (the REPL, IJulia, etcetera) to make it easier to work interactively with Julia, especially for beginners.

In particular, SoftGlobalScope provides a function softscope that can transform Julia code from using the default "hard" (local) scoping rules to simpler "soft" scoping rules in global scope, and a function softscope_include_string that can evaluate a whole string (similar to include_string) using these rules.

For example, if s is a global variable in the current module (e.g. Main), the following code is an error in Julia 1.0:

for i = 1:10
    s += i     # declares a new local variable `s`
end

Instead, you can transform the expression using softscope to automatically insert the necessary global keyword:

julia> softscope(Main, :(for i = 1:10
           s += i
       end))
:(for i = 1:10
      #= REPL[3]:2 =#
      global s += i
  end)

You can then execute the statement with eval. Alternatively, you can decorate the expression with the @softscope macro:

julia> s = 0;

julia> @softscope for i = 1:10
           s += i
       end

julia> s
55

This macro should only be used in the global scope (e.g., via the REPL); using this macro within a function is likely to lead to unintended consequences.

You can execute an entire sequence of statements using "soft" global scoping rules via softscope_include_string:

julia> softscope_include_string(Main, """
       s = 0
       for i = 1:10
           s += i
       end
       s
       """)
55

(This function works like include_string, returning the value of the last evaluated expression.)

On Julia 0.6, softscope is the identity and softscope_include_string is equivalent to include_string, since the global keyword is not needed there.

SoftGlobalScope._add_linenum

SoftGlobalScope._softscope

SoftGlobalScope.add_linenum

SoftGlobalScope.assignments

SoftGlobalScope.calls

SoftGlobalScope.eval

SoftGlobalScope.include

SoftGlobalScope.localassignment

SoftGlobalScope.localvars

SoftGlobalScope.shift_linenum

SoftGlobalScope.softscope — Function.

softscope(m::Module, ast)

Transform the abstract syntax tree ast (a quoted Julia expression) to use "soft" scoping rules for the global variables defined in m, returning the new expression.

SoftGlobalScope.softscope_include_string — Function.

softscope_include_string(m::Module, code::AbstractString, filename::AbstractString="string")

Like include_string, but evaluates code using "soft" scoping rules for the global variables defined in m.

SoftGlobalScope.toplevel_only