Docstrings

FileIO.@format_str

FileIO.DataFormat — Type.

DataFormat{sym}() indicates a known binary or text format of kind sym, where sym is always a symbol. For example, a .csv file might have DataFormat{:CSV}().

An easy way to write DataFormat{:CSV} is format"CSV".

FileIO.File — Type.

File(fmt, filename) indicates that filename is a file of known DataFormat fmt. For example, File{fmtpng}(filename) would indicate a PNG file.

FileIO.FileIO — Module.

FileIO API (brief summary, see individual functions for more detail):

format"PNG": specifies a particular defined format

File{fmt} and Stream{fmt}: types of objects that declare that a resource has a particular format fmt

load([filename|stream]): read data in formatted file, inferring the format

load(File(format"PNG",filename)): specify the format manually

loadstreaming([filename|stream]): similar to load, except that it returns an object that can be read from

save(filename, data...) for similar operations involving saving data

savestreaming([filename|stream]): similar to save, except that it returns an object that can be written to

io = open(f::File, args...) opens a file

io = stream(s::Stream) returns the IOStream from the query object s

query([filename|stream]): attempt to infer the format of filename

unknown(q) returns true if a query can't be resolved

skipmagic(io, fmt) sets the position of io to just after the magic bytes

magic(fmt) returns the magic bytes for format fmt

info(fmt) returns (magic, extensions) for format fmt

add_format(fmt, magic, extension): register a new format

add_loader(fmt, :Package): indicate that Package supports loading files of type fmt

add_saver(fmt, :Package): indicate that Package supports saving files of type fmt

FileIO.Formatted

FileIO.LOAD

FileIO.Linux

FileIO.LoaderError — Type.

LoaderError should be thrown when loader library code fails, and other libraries should be given the chance to recover from the error. Reports the library name and an error message: LoaderError("ImageMagick", "Foo not available")

FileIO.MimeWriter

FileIO.NotInstalledError — Type.

NotInstalledError should be thrown when a library is currently not installed.

FileIO.OS

FileIO.OSX

FileIO.SAVE

FileIO.Stream — Type.

Stream(fmt, io, [filename]) indicates that the stream io is written in known Format. For example, Stream{PNG}(io) would indicate PNG format. If known, the optional filename argument can be used to improve error messages, etc.

FileIO.Unix

FileIO.UnknownFormat — Type.

UnknownFormat gets thrown when FileIO can't recognize the format of a file.

FileIO.Windows

FileIO.WriterError — Type.

WriterError should be thrown when writer library code fails, and other libraries should be given the chance to recover from the error. Reports the library name and an error message: WriterError("ImageMagick", "Foo not available")

FileIO._findmod

FileIO.add_extension

FileIO.add_format — Function.

add_format(fmt, magic, extention) registers a new DataFormat. For example:

add_format(format"PNG", (UInt8[0x4d,0x4d,0x00,0x2b], UInt8[0x49,0x49,0x2a,0x00]), [".tiff", ".tif"])
add_format(format"PNG", [0x89,0x50,0x4e,0x47,0x0d,0x0a,0x1a,0x0a], ".png")
add_format(format"NRRD", "NRRD", [".nrrd",".nhdr"])

Note that extensions, magic numbers, and format-identifiers are case-sensitive.

FileIO.add_loader — Function.

add_loader(fmt, :Package) triggers using Package before loading format fmt

FileIO.add_loadsave

FileIO.add_saver — Function.

add_saver(fmt, :Package) triggers using Package before saving format fmt

FileIO.applicable_loaders

FileIO.applicable_savers

FileIO.applies_to_os

FileIO.canonicalize_magic

FileIO.checked_import

FileIO.del_extension

FileIO.del_format — Function.

del_format(fmt::DataFormat) deletes fmt from the format registry.

FileIO.del_magic

FileIO.detect_bedgraph

FileIO.detect_noometiff

FileIO.detect_ometiff

FileIO.detect_rdata

FileIO.detect_rdata_single

FileIO.detect_stlascii

FileIO.detect_stlbinary

FileIO.detectavi

FileIO.detecthdf5

FileIO.detecttiff

FileIO.detectwav

FileIO.eval

FileIO.ext2sym

FileIO.file!

FileIO.file_extension — Function.

file_extension(file) returns the file extension associated with File file.

file_extension(file) returns a nullable-string for the file extension associated with Stream stream.

FileIO.filename — Function.

filename(file) returns the filename associated with File file.

filename(stream) returns a string of the filename associated with Stream stream, or nothing if there is no file associated.

FileIO.getmodule

FileIO.h5magic

FileIO.handle_current_error — Function.

Handles error as soon as they get thrown while doing IO

FileIO.handle_error

FileIO.handle_exceptions — Function.

Handles a list of thrown errors after no IO library was found working

FileIO.has_method_from

FileIO.hasfunction

FileIO.hasmagic

FileIO.include

FileIO.info — Function.

info(fmt) returns the magic bytes/extension information for DataFormat fmt.

FileIO.is_installed

FileIO.iter_eq

FileIO.lensym

FileIO.load — Function.

load(filename) loads the contents of a formatted file, trying to infer

the format from filename and/or magic bytes in the file.

load(strm) loads from an IOStream or similar object. In this case,

there is no filename extension, so we rely on the magic bytes for format identification.

load(File(format"PNG", filename)) specifies the format directly, and bypasses inference.

load(Stream(format"PNG", io)) specifies the format directly, and bypasses inference.

load(f; options...) passes keyword arguments on to the loader.

FileIO.loadstreaming — Function.

Some packages may implement a streaming API, where the contents of the file can be read in chunks and processed, rather than all at once. Reading from these higher-level streams should return a formatted object, like an image or chunk of video or audio.

loadstreaming(filename) loads the contents of a formatted file, trying to infer

the format from filename and/or magic bytes in the file. It returns a streaming type that can be read from in chunks, rather than loading the whole contents all at once

loadstreaming(strm) loads the stream from an IOStream or similar object.

In this case, there is no filename extension, so we rely on the magic bytes for format identification.

loadstreaming(File(format"WAV",filename)) specifies the format directly, and

bypasses inference.

loadstreaming(Stream(format"WAV", io)) specifies the format directly, and

bypasses inference.

loadstreaming(f; options...) passes keyword arguments on to the loader.

FileIO.magic — Function.

magic(fmt) returns the magic bytes of format fmt

FileIO.magic_cmp

FileIO.magic_equal

FileIO.magic_func

FileIO.magic_list

FileIO.metadata

FileIO.query — Function.

query(filename) returns a File object with information about the format inferred from the file's extension and/or magic bytes.

query(io, [filename]) returns a Stream object with information about the format inferred from the magic bytes.

FileIO.save — Function.

save(filename, data...) saves the contents of a formatted file,

trying to infer the format from filename.

save(Stream(format"PNG",io), data...) specifies the format directly, and bypasses inference.

save(File(format"PNG",filename), data...) specifies the format directly, and bypasses inference.

save(f, data...; options...) passes keyword arguments on to the saver.

FileIO.savestreaming — Function.

Some packages may implement a streaming API, where the contents of the file can be written in chunks, rather than all at once. These higher-level streams should accept formatted objects, like an image or chunk of video or audio.

savestreaming(filename, data...) saves the contents of a formatted file,

trying to infer the format from filename.

savestreaming(File(format"WAV",filename)) specifies the format directly, and

bypasses inference.

savestreaming(Stream(format"WAV", io)) specifies the format directly, and

bypasses inference.

savestreaming(f, data...; options...) passes keyword arguments on to the saver.

FileIO.seekable

FileIO.skipmagic — Function.

skipmagic(s) sets the position of Stream s to be just after the magic bytes. For a plain IO object, you can use skipmagic(io, fmt).

FileIO.split_predicates

FileIO.stream — Function.

stream(s) returns the stream associated with Stream s

FileIO.sym2info

FileIO.sym2loader

FileIO.sym2saver

FileIO.tiff_magic

FileIO.topimport

FileIO.unknown — Function.

unknown(f) returns true if the format of f is unknown.

FileIO.unknown_df — Type.

DataFormat{sym}() indicates a known binary or text format of kind sym, where sym is always a symbol. For example, a .csv file might have DataFormat{:CSV}().

An easy way to write DataFormat{:CSV} is format"CSV".