Docstrings

GlobalSearchRegression.AVAILABLE_CRITERIA

GlobalSearchRegression.AVAILABLE_METHODS

GlobalSearchRegression.AVAILABLE_VARIABLES

GlobalSearchRegression.CRITERIA_DEFAULT

GlobalSearchRegression.CRITERIA_DEFAULT_INSAMPLE

GlobalSearchRegression.CRITERIA_DEFAULT_OUTSAMPLE

GlobalSearchRegression.CSV_DEFAULT

GlobalSearchRegression.CSV_DUPLICATED_PARAMETERS

GlobalSearchRegression.DATANAMES_REQUIRED

GlobalSearchRegression.DATA_NOT_DEFINED

GlobalSearchRegression.EQUATION_GENERAL_INFORMATION

GlobalSearchRegression.GSRegResult

GlobalSearchRegression.GlobalSearchRegression

GlobalSearchRegression.INDEX

GlobalSearchRegression.INSAMPLE_MIN_SIZE

GlobalSearchRegression.INTERCEPT_DEFAULT

GlobalSearchRegression.METHOD_DEFAULT

GlobalSearchRegression.METHOD_INVALID

GlobalSearchRegression.MODEL_AVG_DEFAULT

GlobalSearchRegression.NO_ENOUGH_OBSERVATIONS

GlobalSearchRegression.ON_MESSAGE_DEFAULT

GlobalSearchRegression.ORDER

GlobalSearchRegression.ORDER_RESULTS_DEFAULT

GlobalSearchRegression.OUTSAMPLE_DEFAULT

GlobalSearchRegression.OUTSAMPLE_HIGHER_VALUE

GlobalSearchRegression.OUTSAMPLE_LOWER_VALUE

GlobalSearchRegression.OUTSAMPLE_MISMATCH

GlobalSearchRegression.PARALLEL_DEFAULT

GlobalSearchRegression.RESIDUALTEST_EXCEPTION

GlobalSearchRegression.RESIDUAL_TESTS_CROSS

GlobalSearchRegression.RESIDUAL_TESTS_TIME

GlobalSearchRegression.RESIDUAL_TEST_DEFAULT

GlobalSearchRegression.SELECTED_VARIABLES_DOES_NOT_EXISTS

GlobalSearchRegression.SUMMARY_DEFAULT

GlobalSearchRegression.TIME_DEFAULT

GlobalSearchRegression.TIME_VARIABLE_INEXISTENT

GlobalSearchRegression.TTEST_DEFAULT

GlobalSearchRegression.VARIABLES_NOT_OR_VALID_OR_NOT_DEFINED

GlobalSearchRegression.VECTOR_OPERATION_DEFAULT

GlobalSearchRegression.WEIGHT

GlobalSearchRegression.WRONG_DATASET_DATATYPE

GlobalSearchRegression.convert_if_is_dataframe_to_array

GlobalSearchRegression.convert_if_is_tuple_to_array

GlobalSearchRegression.datanames_strarr_to_symarr!

GlobalSearchRegression.equation_str_to_strarr

GlobalSearchRegression.equation_strarr_to_symarr

GlobalSearchRegression.eval

GlobalSearchRegression.export_csv — Function.

Exports main results with headers to file

GlobalSearchRegression.filter_data_valid_columns

GlobalSearchRegression.filter_rows_with_empty_values

GlobalSearchRegression.get_data_column_pos

GlobalSearchRegression.get_data_position — Function.

Returns the position of the header value based on this structure. - Index - Covariates * b * bstd * T-test - Equation general information merged with criteria user-defined options. - Order from user combined criteria - Weight

GlobalSearchRegression.get_datanames

GlobalSearchRegression.get_result_header — Function.

Constructs the header for results based in getdataposition orders.

GlobalSearchRegression.get_selected_cols — Function.

Returns selected appropiate covariates for each iteration

GlobalSearchRegression.gsreg

GlobalSearchRegression.gsreg_proc_result!

GlobalSearchRegression.gsreg_single_proc_result!

GlobalSearchRegression.gsregsortrows

GlobalSearchRegression.in_vector

GlobalSearchRegression.include

GlobalSearchRegression.proc!

GlobalSearchRegression.sort_data_by_time

GlobalSearchRegression.to_dict

GlobalSearchRegression.to_string