Docstrings

Schemata.CATEGORICAL

Schemata.ColumnSchema

Schemata.REQUIRED

Schemata.Schema

Schemata.Schemata

Schemata.TableSchema

Schemata.UNIQUE

Schemata.construct_intrarow_constraints

Schemata.convert_args_types

Schemata.determine_eltype — Function.

Determine ColumnSchema.eltyp.

Schemata.determine_validvalues — Function.

Determine ColumnSchema.validvalues from vv and ColumSchema.eltyp (cseltyp).

Schemata.determine_vv — Function.

Returns: An instance of ColumnSchema.valid_values.

vv could be a:

DataType. E.g., "Int64"

Range of a Base type. E.g., "1:10"

Range of a non-Base type, represented with a tuple: "(val1, val2)" or "(val1, stepsize, val2)"

Used if valid_values is the same as eltyp, and vv != datatype.

Schemata.diagnose — Function.

Returns: Collection containing ways in which the table does not comply with the schema.

Default collection is a vector of tuples. If DataFrames is defined the collection is a DataFrame.

Example result:

:entity :id :issue col patientid Incorrect data type (String) col patientid Missing data not allowed col patientid Values are not unique col gender Invalid values ('d') table mytable Primary key not unique

Schemata.diagnose_column! — Function.

Append table-level issues into issues.

Schemata.diagnose_table! — Function.

Modified: issues

Schemata.enforce_schema — Function.

Returns: table, issues

The table is as compliant as possible with the schema. If the table is completely compliant with the schema, then the issues table has 0 rows. Otherwise the issues table contains the ways in which the output table does not comply with the schema.

Schemata.eval

Schemata.get_datatype

Schemata.include

Schemata.init_compliant_data — Function.

Returns: A table with unpopulated columns with name, type, length and order matching the table schema.

Schemata.insert_column! — Function.

Insert a column into the table schema at position n.

Schemata.parse_as_type

Schemata.parse_nonbase_range — Function.

Used to represent ranges for non-Base types.

The format is: "(start, stop)" or "(start, stepsize, stop)"

If vv has 2 entries, these represent the end-points of the range; i.e., start:stop. If vv has 3 entries, the middle entry represents the step size; i.e., start:stepsize:stop.

Example: (2017-10-01 09:00, 2017-12-08 23:00), where the entries have type TimeZones.ZonedDateTime.

For Base types a stringified range will work, E.g., eval(Meta.parse("1:10")) will return the range 1:10. For non-Base types this approach will fail. E.g., eval(Meta.parse("2017-10-01 09:00:2017-12-08 23:00")) will fail.

Schemata.readschema

Schemata.value_is_valid