Docstrings

Expectations.Expectation — Type.

Abstract type for all expectations.

Expectations.Expectations

Expectations.ExplicitQuadratureAlgorithm — Type.

Abstract type for quadrature algorithms with user-defined nodes (e.g., trapezoidal integration).

Expectations.FiniteDiscrete — Type.

A dot product of a (finite) PDF vector and a finite set of transformed nodes.

Expectations.Gaussian — Type.

Gaussian quadrature. See specific methods for what precise algorithm is used (e.g., Gauss-Legendre, Gauss-Hermite, etc.)

Expectations.IterableExpectation — Type.

Expectations which are paramterized by a vector of nodes (e.g., a discretized support) and corresponding quadrature weights.

Expectations.QuadratureAlgorithm — Type.

Abstract type for quadrature algorithms without user-defined nodes (e.g., Gaussian quadrature.)

Expectations.QuantileRange — Type.

A custom quadrature scheme written by Spencer Lyon as part of the QuantEcon.jl library. Used with permission.

For detailed information, see: https://github.com/QuantEcon/QuantEcon.jl/blob/be0a32ec17d1f5b04ed8f2e52604c70c69f416b2/src/quad.jl#L918.

Expectations.Trapezoidal — Type.

Trapezoidal integration.

Expectations._expectation — Function.

function _expectation(dist::DiscreteUnivariateDistribution, alg::Type{FiniteDiscrete}; kwargs...)

Auxiliary constructor for an IterableExpectation object.

function _expectation(dist::ContinuousUnivariateDistribution, alg::Type{Gaussian}; n = 500, kwargs...)

Implements Gauss-Legendre quadrature for continuous univariate distributions for which no specialized method exists.

function _expectation(dist::ContinuousUnivariateDistribution, alg::Type{QuantileRange}; n::Int = 50, q0::Real = 0.001, qN::Real = 0.999 kwargs...)

Implementation of the qnwdist() quadrature scheme written by Spencer Lyon (PhD. NYU), as part of the QuantEcon.jl library. Used with permission. For further details, see: https://github.com/QuantEcon/QuantEcon.jl/blob/be0a32ec17d1f5b04ed8f2e52604c70c69f416b2/src/quad.jl#L892.

function _expectation(dist::Normal, alg::Type{Gaussian}; n = 30, kwargs...)

Implements Gauss-Hermite quadrature for normal distributions.

function _expectation(dist::LogNormal, alg::Type{Gaussian}; n = 30, kwargs...)

Implements Gauss-Hermite quadrature for lognormal distributions.

function _expectation(dist::Beta, alg::Type{Gaussian}; n = 32, kwargs...)

Implements Gauss-Jacobi quadrature for beta distributions.

function _expectation(dist::Exponential, alg::Type{Gaussian}; n = 32, kwargs...)

Implements Gauss-Laguerre quadrature for Exponential distributions.

function _expectation(dist::Gamma, alg::Type{Gaussian}; n = 32, kwargs...)

Implements Gauss-Laguerre quadrature for Gamma distributions.

function _expectation(dist, nodes::AbstractArray, alg::Type{Trapezoidal}; kwargs...)

Implements trapezoidal integration for general distributions with user-defined nodes.

@compat function _expectation(dist, nodes::AbstractRange, alg::Type{Trapezoidal}; kwargs...)

Overloads trapezoidal integration for cases where the user-defined grids are regular.

Expectations.eval

Expectations.expectation — Function.

expectation(dist::DiscreteUnivariateDistribution, alg::Type{FiniteDiscrete} = FiniteDiscrete; kwargs...) = _expectation(dist, alg; kwargs...)

Dispatcher for (finite) discrete univariate expectations.

expectation(dist::ContinuousUnivariateDistribution, alg::Type{<:QuadratureAlgorithm} = Gaussian; kwargs...) = _expectation(dist, alg; kwargs...)

Dispatcher for continuous univariate expectations.

expectation(dist::ContinuousUnivariateDistribution, nodes, alg::Type{<:ExplicitQuadratureAlgorithm} = Trapezoidal; kwargs...)  = _expectation(dist, nodes, alg; kwargs...)

Dispatcher for distributions with user-defined nodes.

expectation(f::Function, dist::DiscreteUnivariateDistribution, alg::Type{FiniteDiscrete} = FiniteDiscrete; kwargs...) = expectation(dist, alg; kwargs...)(f)

Convenience function for (finite) discrete univariate distributions.

expectation(f::Function, dist::ContinuousUnivariateDistribution, alg::Type{<:QuadratureAlgorithm} = Gaussian; kwargs...) = expectation(dist, alg; kwargs...)(f)

Convenience function for continuous univariate distributions.

expectation(f::Function, dist::ContinuousUnivariateDistribution, nodes::AbstractArray, alg::Type{<:ExplicitQuadratureAlgorithm} = Trapezoidal; kwargs...) = expectation(dist, nodes, alg; kwargs...)(f)

Convenience function for continuous univariate distributions with user-supplied nodes.

Expectations.include

Expectations.nodes — Function.

nodes(e::IterableExpectation)

Returns the nodes of an IterableExpectation.

Expectations.weights — Function.

weights(e::IterableExpectation)

Returns the weights of an IterableExpectation.