Docstrings

Laplacians.ApproxCholPQ — Type.

An approximate priority queue. Items are bundled together into doubly-linked lists with all approximately the same key. minlist is the min list we know to be non-empty. It should always be a lower bound. keyMap maps keys to lists

Laplacians.ApproxCholPQElem

Laplacians.ApproxCholParams — Type.

params = ApproxCholParams(order, output)

order can be one of

:deg (by degree, adaptive),

:wdeg (by original wted degree, nonadaptive),

:given

Laplacians.AugTreeParams

Laplacians.AugTreeParamsOld

Laplacians.ErdosRenyi — Function.

graph = ErdosRenyi(n::Integer, m::Integer; ver=Vcur)

Generate a random graph on n vertices with m edges. The actual number of edges will probably be smaller, as we sample with replacement

Laplacians.ErdosRenyiCluster — Function.

graph = ErdosRenyiCluster(n::Integer, k::Integer; ver=Vcur)

Generate an ER graph with average degree k, and then return the largest component. Will probably have fewer than n vertices. If you want to add a tree to bring it back to n, try ErdosRenyiClusterFix.

Laplacians.ErdosRenyiClusterFix — Function.

graph = ErdosRenyiClusterFix(n::Integer, k::Integer; ver=Vcur)

Like an Erdos-Renyi cluster, but add back a tree so it has n vertices

Laplacians.ErdosRenyiClusterFix_ijv

Laplacians.ErdosRenyiCluster_ijv

Laplacians.ErdosRenyi_ijv

Laplacians.FastSampler — Type.

s = FastSampler(p)

Laplacians.HeapEntry

Laplacians.IJV — Type.

ijv = IJV(A::SparseMatrixCSC)

Convert a sparse matrix to an IJV.

Laplacians.IJVS

Laplacians.IJVind

Laplacians.IJVindGraph

Laplacians.IJVindList

Laplacians.KMPLapPrecon

Laplacians.KMPLapPreconSub

Laplacians.KMPLapSolver — Function.

lapSolver = KMPLapSolver(A; verbose, tol, maxits, maxtime, pcgIts, params::KMPParams)

Solves linear equations in the Laplacian of graph with adjacency matrix A.

Based on the paper "Approaching optimality for solving SDD systems" by Koutis, Miller, and Peng, <i>SIAM Journal on Computing</i>, 2014.

Laplacians.KMPLapSolver1

Laplacians.KMPParams — Type.

Parameters for the KMP solver

Laplacians.KMPSDDMSolver — Function.

sddmSolver = KMPSDDMSolver(mat; verbose, tol, maxits, maxtime, pcgIts, params::KMPParams)

Solves linear equations in symmetric, diagonally dominant matrices with non-positive off-diagonals. Based on the paper "Approaching optimality for solving SDD systems" by Koutis, Miller, and Peng, <i>SIAM Journal on Computing</i>, 2014.

Laplacians.LDLinv — Type.

LDLinv contains the information needed to solve the Laplacian systems. It does it by applying Linv, then Dinv, then Linv (transpose). But, it is specially constructed for this particular solver. It does not explicitly make the matrix triangular. Rather, col[i] is the name of the ith col to be eliminated

Laplacians.LDLsolver

Laplacians.LLMatOrd

Laplacians.LLcol

Laplacians.LLmatp — Type.

LLmatp is the data structure used to maintain the matrix during elimination. It stores the elements in each column in a singly linked list (only next ptrs) Each element is an LLp (linked list pointer). The head of each column is pointed to by cols.

We probably can get rid of degs - as it is only used to store initial degrees.

Laplacians.LLord

Laplacians.LLp — Type.

LLp elements are all in the same column. row tells us the row, and val is the entry. val is set to zero for some edges that we should remove. next gives the next in the column. It points to itself to terminate. reverse is the index into lles of the other copy of this edge, since every edge is stored twice as we do not know the order of elimination in advance.

Laplacians.LapSolvers — Constant.

A list containing Laplacian linear system solvers. They take in an adjacency matrix plus tol, maxits and maxtime parameters.

Laplacians.Laplacians — Module.

A package for graph computations related to graph Laplacians

Graphs are represented by sparse adjacency matrices, etc.

Laplacians.LinkedListStorage

Laplacians.RootedTree

Laplacians.SDDMSolvers — Constant.

A list containing SDDM linear system solvers. They take in a SDDM matrix plus tol, maxits and maxtime parameters.

Laplacians.Sampler

Laplacians.SolverTest — Type.

SolverTest(solver, name)

Encloses a solver with its name, so that we can compare it in tests

Laplacians.SqLinOp

Laplacians.V06

Laplacians.V07

Laplacians.Vcur

Laplacians.__init__

Laplacians.addToGPrime — Function.

Add a new vertex to GPrime

Laplacians.adj — Function.

a,d = adj(sddm)

Create an adjacency matrix and a diagonal vector from an SDD M-matrix. That is, from a Laplacian with added diagonal weights

Laplacians.akpw — Function.

tree = akpw(graph; ver=0)

Computes a low stretch spanning tree of graph, and returns it as a graph. The default version is 0. In event of emergency, one can try ver=2. It is usually slower, but might have slightly better stretch.

Laplacians.akpwSub2

Laplacians.akpwSub5

Laplacians.akpwU — Function.

tree = akpwU(graph)

Computes a low stretch spanning tree of an unweighted graph, and returns it as a graph.

Laplacians.akpwUsub

Laplacians.approxChol

Laplacians.approxCholLap

Laplacians.approxCholPQDec! — Function.

Decrement the key of element i
This could crash if i exceeds the maxkey

Laplacians.approxCholPQInc! — Function.

Increment the key of element i
This could crash if i exceeds the maxkey

Laplacians.approxCholPQMove!

Laplacians.approxCholPQPop!

Laplacians.approxCholSDDM

Laplacians.approxQual — Function.

eps = approxQual(graph1, graph2; tol=1e-5, verbose=false)

Computes the eps for which graph1 and graph2 are eps approximations of each other. That is, L1 <= (1+eps) L2, and vice versa.

It is randomized, so you might want to run it again if you don't trust the answers.

Laplacians.approxchol_lap — Function.

solver = approxchol_lap(a); x = solver(b);
solver = approxchol_lap(a; tol::Real=1e-6, maxits=1000, maxtime=Inf, verbose=false, pcgIts=Int[], params=ApproxCholParams())

A heuristic by Daniel Spielman inspired by the linear system solver in https://arxiv.org/abs/1605.02353 by Rasmus Kyng and Sushant Sachdeva. Whereas that paper eliminates vertices one at a time, this eliminates edges one at a time. It is probably possible to analyze it. The ApproxCholParams let you choose one of three orderings to perform the elimination.

ApproxCholParams(:given) - in the order given. This is the fastest for construction the preconditioner, but the slowest solve.

ApproxCholParams(:deg) - always eliminate the node of lowest degree. This is the slowest build, but the fastest solve.

ApproxCholParams(:wdeg) - go by a perturbed order of wted degree.

For more info, see http://danspielman.github.io/Laplacians.jl/latest/usingSolvers/index.html

Laplacians.approxchol_lap1

Laplacians.approxchol_lapChol — Function.

This variation of approxChol creates a cholesky factor to do the elimination. It has not yet been optimized, and does not yet make the cholesky factor lower triangular

Laplacians.approxchol_lapGiven

Laplacians.approxchol_lapGreedy

Laplacians.approxchol_lapWdeg

Laplacians.approxchol_sddm — Function.

solver = approxchol_sddm(sddm); x = solver(b);
solver = approxchol_sddm(sddm; tol=1e-6, maxits=1000, maxtime=Inf, verbose=false, pcgIts=Int[], params=ApproxCholParams())

Solves sddm systems by wrapping approxchol_lap. Not yet optimized directly for sddm.

For more info, see http://danspielman.github.io/Laplacians.jl/latest/usingSolvers/index.html

Laplacians.apr — Function.

Computes an approximate page rank vector from a starting set s, an alpha and an epsilon The algorithm follows the Anderson,Chung,Lang paper and Dan Spielman's lecture notes

Laplacians.ask_cholmod — Function.

nnzL, flops = ask_cholmod(mat)

Estimate the number of nonzeros in the cholfact factorization of mat, along with the number of flops needed to compute it. Does this through a call to the analyze routine of cholmod. Note that this is much faster than actually computing the factorization

Laplacians.augTreeFactor

Laplacians.augTreeLap — Function.

solver = augTreeLap(A; tol::Real=1e-6, maxits=Inf, maxtime=Inf, verbose=false, pcgIts=Int[], params=AugTreeParams())

An "augmented spanning tree" solver for Laplacians. It works by adding edges to a low stretch spanning tree. It calls augTreePrecon to form the preconditioner. params has entries

params.treeAlg default to akpw

params.opt if true, it interacts with cholmod to choose a good number of edges to add back. If false, it adds back 2*sqrt(n).

Laplacians.augTreeLap1

Laplacians.augTreeLapPrecon — Function.

pre = augTreeLapPrecon{Tv,Ti}(A; params=AugTreeParams())

This is an augmented spanning tree preconditioner for Laplacians. It takes as optional input a tree growing algorithm. It adds back 2sqrt(n) edges via augmentTree: the sqrt(n) of highest stretch and another sqrt(n) sampled according to stretch. For most purposes, one should directly call augTreeLapSolver.

Laplacians.augTreePrecon — Function.

pre = augTreePrecon{Tv,Ti}(ddmat::SparseMatrixCSC{Tv,Ti}; params=AugTreeParams())

This is an augmented spanning tree preconditioner for diagonally dominant linear systems. It takes as optional input a tree growing algorithm. It adds back 2sqrt(n) edges via augmentTree: the sqrt(n) of highest stretch and another sqrt(n) sampled according to stretch. For most purposes, one should directly call augTreeSolver.

Laplacians.augTreeSddm — Function.

solver = augTreeSddm(sddm; tol::Real=1e-6, maxits=Inf, maxtime=Inf, verbose=false, pcgIts=Int[],  params=AugTreeParams())

An "augmented spanning tree" solver for positive definite diagonally dominant matrices. It works by adding edges to a low stretch spanning tree. It calls augTreePrecon to form the preconditioner. params has entries

params.treeAlg default to akpw

params.opt if true, it interacts with cholmod to choose a good number of edges to add back. If false, it adds back 2*sqrt(n).

Laplacians.augmentTree — Function.

B = augmentTree{Tv,Ti}(tree, A, k)

Takes as input a tree and an adjacency matrix of a graph. It then computes the stretch of every edge of the graph wrt the tree. It then adds back the k edges of highest stretch, and k edges sampled according to stretch.

This is the old alg. We now recommend using augmentTreeOpt.

Laplacians.augmentTreeOpt — Function.

B = augmentTreeOpt{Tv,Ti}(tree, A, params)

Takes as input a tree and an adjacency matrix of a graph. It then computes the stretch of every edge of the graph wrt the tree. It uses cholmod to decide how many edge to add back, shooting for nnzLfac times n entries in the factored augmented tree, with a number of flops to factor equal to nnz(a)*flopsfac. The edges to add back are then choen at random.

Laplacians.axpy2!

Laplacians.backIndices — Function.

Computes the back indices in a graph in O(M+N). works if for every edge (u,v), (v,u) is also in the graph

Same as the above, but now the graph is in adjacency list form

Laplacians.backSolve

Laplacians.backward!

Laplacians.bfsFromSeed

Laplacians.bfsOrder

Laplacians.biggestComp — Function.

Return the biggest component in a graph, as a graph

Laplacians.blockSample — Function.

s = blockSample(p; k = length(p))

Compute numbers sampled with probability proportional to the vector p. They are returned in order. If need be, they can be permuted with randperm.

Laplacians.blockSolver — Function.

Apply the ith solver on the ith component

Laplacians.bzbeta!

Laplacians.cbt_ijv

Laplacians.cg — Function.

x = cg(mat, b; tol, maxits, maxtime, verbose, pcgIts)

solves a symmetric linear system mat x = b.

Arguments

tol is set to 1e-6 by default,

maxits defaults to Inf

maxtime defaults to Inf. It measures seconds.

verbose defaults to false

pcgIts is an array for returning the number of pcgIterations. Default is length 0, in which case nothing is returned.

Laplacians.cgBLAS

Laplacians.cgLapSolver — Function.

x = cgLapSolver(A::AbstractMatrix; tol::Real=1e-6, maxits=Inf, maxtime=Inf, verbose=false, pcgIts=Int[])

Create a solver that uses cg to solve Laplacian systems in the laplacian of A. This just exists to satisfy our interface. It does nothing more than create the Laplacian and call cg on each connected component.

Laplacians.cgLapSolver1

Laplacians.cgSlow

Laplacians.cgSolver — Function.

x = cgSolver(mat; tol, maxits, maxtime, verbose, pcgIts)

creates a solver for a PSD system mat. The parameters are as described in cg.

Laplacians.chimera — Function.

graph = chimera(n::Integer; verbose=false, ver=Vcur)

Builds a chimeric graph on n vertices. The components come from pureRandomGraph, connected by joinGraphs, productGraph and generalizedNecklace

graph = chimera(n::Integer, k::Integer; verbose=false, ver=Vcur)

Builds the kth chimeric graph on n vertices. It does this by resetting the random number generator seed. It should captute the state of the generator before that and then return it, but it does not yet.

Laplacians.chimera_ijv

Laplacians.cholLap

Laplacians.cholSDDM

Laplacians.chol_lap — Function.

solver = chol_lap(A::AbstractArray)

Uses Cholesky Factorization to solve systems in Laplacians.

Laplacians.chol_sddm — Function.

solveSDDM = chol_sddm(sddm::AbstractMatrix; tol, maxits, maxtime, verbose, pcgIts=Int[])

This functions wraps cholfact so that it satsfies our interface. It ignores all the keyword arguments.

Laplacians.cholmod_perm — Function.

p = cholmod_perm(mat)

Return the permutation that cholmod would apply.

Laplacians.cluster!

Laplacians.combineMultiG

Laplacians.compConductance — Function.

Returns the quality of the cut for a given graph and a given cut set s. the result will be |outgoing edges| / min(|vertices in set|, |N - vertices in set|)

Laplacians.compDepth

Laplacians.compGraph

Laplacians.compGraphU

Laplacians.compStretches

Laplacians.comp_stretches — Function.

Compute the stretched of every edge in mat with respect to the tree tree. Returns the answer as a sparse matrix with the same nonzero structure as mat. Assumes that mat is symmetric. tree should be the adjacency matrix of a spanning tree.

Laplacians.completeBinaryTree

Laplacians.completeGraph

Laplacians.complete_binary_tree — Function.

graph = completeBinaryTree(n::Int64)

The complete binary tree on n vertices

Laplacians.complete_graph — Function.

graph = complete_graph(n)

Laplacians.complete_graph_ijv — Function.

ijv = complete_graph_ijv(n)

Laplacians.components — Function.

Computes the connected components of a graph. Returns them as a vector of length equal to the number of vertices. The vector numbers the components from 1 through the maximum number. For example,

gr = ErdosRenyi(10,11)
c = components(gr)

10-element Array{Int64,1}:
 1
 1
 1
 1
 2
 1
 1
 1
 3
 2

Laplacians.compress

Laplacians.compressCol!

Laplacians.compressIndices!

Laplacians.computeCN

Laplacians.condNumber — Function.

cn = condNumber(a, ldli; verbose=false)

Given an adjacency matrix a and an ldli computed by approxChol, this computes the condition number.

Laplacians.conditionNumber — Function.

kapps = conditionNumber(graph1, graph2; tol=1e-5, verbose=false)

Computes the relative condition number of graph1 and graph2.

It is randomized, so you might want to run it again if you don't trust the answers.

kappa = conditionNumber(graph, precon; tol=1e-5, verbose=false)

Computes the relative condition number of graph and a preconditioning function.

It is randomized, so you might want to run it again if you don't trust the answers.

Laplacians.defaultKMPParams

Laplacians.deg

Laplacians.dfsOrder

Laplacians.diagmat — Function.

d = diagmat(a)

Returns the diagonal weighted degree matrix(as a sparse matrix) of a graph

Laplacians.dijkstraFromSeed

Laplacians.dirEdgeVertexMat — Function.

The signed edge-vertex adjacency matrix

Laplacians.disjoin — Function.

graph = disjoin(a,b)

Create a disjoint union of graphs a and b, with no edges between them.

Laplacians.disjoin!

Laplacians.dumbRefineCut — Function.

Modify a cluster by passing through all the vertices exactly once and 
adding/removing them based on the value of (Deg_external - Deg_Internal).

Laplacians.edgeVertexMat — Function.

U = edgeVertexMat(a)

The signed edge-vertex adjacency matrix

Laplacians.element

Laplacians.elimDeg12

Laplacians.elimDeg2Node

Laplacians.elimLeafNode

Laplacians.empty_graph — Function.

ijv = empty_graph(n)

Laplacians.empty_graph_ijv — Function.

ijv = empty_graph_ijv(n)

Laplacians.eval

Laplacians.extendMatrix — Function.

Add a new vertex to a with weights to the other vertices corresponding to diagonal surplus weight.

This is an efficient way of writing [a d; d' 0]

Laplacians.fastPairQueue

Laplacians.fastQueue

Laplacians.findEntries — Function.

Similar to findnz, but also returns 0 entries that have an edge in the sparse matrix

Laplacians.firstn — Function.

b = firstn(a::IJV, n::Integer)

Only keep the first n vertices of a.

Laplacians.flipIndex — Function.

For a symmetric matrix, this gives the correspondance between pairs of entries in an ijv. So, ai[ind] = aj[flip[ind]]. For example,

(ai,aj,av) = findnz(a);
fl = flipIndex(a)
ind = 10
@show backind = fl[10]
@show [ai[ind], aj[ind], av[ind]]
@show [ai[backind], aj[backind], av[backind]];

backind = fl[10] = 4
[ai[ind],aj[ind],av[ind]] = [2.0,4.0,0.7]
[ai[backind],aj[backind],av[backind]] = [4.0,2.0,0.7]

Laplacians.floatGraph — Function.

graph = floatGraph(a::SparseMatrixCSC)

Convert the nonzero entries in a graph to Float64.

Laplacians.forceLap — Function.

la = forceLap(a)

Create a Laplacian matrix from an adjacency matrix. If the input looks like a Laplacian, throw a warning and convert it.

Laplacians.forward!

Laplacians.forwardSolve

Laplacians.generalizedNecklace — Function.

graph = generalizedNecklace(A, H, k::Int64)

Constructs a generalized necklace graph starting with two graphs A and H. The resulting new graph will be constructed by expanding each vertex in H to an instance of A. k random edges will be generated between components. Thus, the resulting graph may have weighted edges.

Laplacians.generalizedRing

Laplacians.generalized_necklace

Laplacians.generalized_ring — Function.

graph = generalized_ring(n, gens)

A generalization of a ring graph. The vertices are integers modulo n. Two are connected if their difference is in gens. For example,

generalized_ring(17, [1 5])

Laplacians.generalized_ring_ijv

Laplacians.getCutSet — Function.

Get the min cut from the source - return all vertices in the cut besides the source

Laplacians.getObound — Function.

Computes the number of edges leaving s

Laplacians.getVolume — Function.

Computes the volume of subset s in an unweighted graph G

Laplacians.get_ll_col

Laplacians.grid2 — Function.

graph = grid2(n::Int64, m::Int64; isotropy=1)

An n-by-m grid graph. iostropy is the weighting on edges in one direction.

Laplacians.grid2_ijv

Laplacians.grid2coords — Function.

graph = grid2coords(n::Int64, m::Int64)
graph = grid2coords(n::Int64)

Coordinates for plotting the vertices of the n-by-m grid graph

Laplacians.grid3 — Function.

graph = grid3(n1, n2, n3)
graph = grid3(n)

An n1-by-n2-by-n3 grid graph.

Laplacians.grid3_ijv

Laplacians.grownGraph

Laplacians.grownGraphD

Laplacians.grown_graph — Function.

graph = grown_graph(n, k; ver=Vcur)

Create a graph on n vertices. For each vertex, give it k edges to randomly chosen prior vertices. This is a variety of a preferential attachment graph.

Laplacians.grown_graph_d — Function.

graph = grown_graph_d(n::Integer, k::Integer; ver=Vcur)

Like a grownGraph, but it forces the edges to all be distinct. It starts out with a k+1 clique on the first k vertices

Laplacians.grown_graph_d_ijv

Laplacians.grown_graph_ijv

Laplacians.hasMore

Laplacians.hyperCube

Laplacians.hypercube — Function.

graph = hyperCube(d::Int64)

The d dimensional hypercube. Has 2^d vertices and d*2^(d-1) edges.

Laplacians.hypercube_ijv

Laplacians.incSize

Laplacians.include

Laplacians.initDictCol! — Function.

initDictCol!(dic, name, typ)

For a dictionary in which each key indexes an array. If dic does not contain an entry of name, create with set to Array(typ,0).

Laplacians.initGPrime — Function.

Initialize GPrime with the set A and edges of type s->u

Laplacians.intHeap

Laplacians.intHeapAdd!

Laplacians.intHeapDown!

Laplacians.intHeapPop!

Laplacians.intHeapSet!

Laplacians.intHeapSort

Laplacians.intHeapUp!

Laplacians.isConnected — Function.

Returns true if graph is connected. Calls components.

Laplacians.johnlind — Function.

Johnson Lindenstrauss effective resistance upperbounding. If retXhat is set to true, returns the vectors necessary for computing the effective resistance bounds instead of the actual bounds.

johnlind(a::SparseMatrixCSC{Tv,Ti}; eps::Tv = 0.5, solver=(la -> augTreeSddm(la,tol=1e-1,maxits=1000,maxtime=10)), retXhat::Bool = false)

Laplacians.join_graphs — Function.

graph = joinGraphs(a, b, k::Integer)

Create a disjoint union of graphs a and b, and then put k random edges between them

Laplacians.join_graphs! — Function.

graph = join_graphs!(a::IJV, b::IJV, k::Integer)

Create a disjoint union of graphs a and b, and then put k random edges between them, merging b into a.

Laplacians.keyMap

Laplacians.kids

Laplacians.kruskal — Function.

(kruskal::SparseMatrixCSC; kind=:max) Uses Kruskal's algorithm to compute a minimum (or maximum) spanning tree. Set kind=:min if you want the min spanning tree. It returns it a a graph

Laplacians.lap — Function.

l = lap(a)

Create a Laplacian matrix from an adjacency matrix. We might want to do this differently, say by enforcing symmetry

Laplacians.lapWrapComponents — Function.

f = lapWrapComponents(solver, a::AbstractArray; tol::Real=1e-6, maxits=Inf, maxtime=Inf, verbose=false, pcgIts=Int[], params...)

Applies a Laplacian solver that satisfies our interface to each connected component of the graph with adjacency matrix a. Passes kwargs on the solver.

Laplacians.lapWrapConnected — Function.

f = lapWrapConnected(sddmSolver, a::AbstractMatrix; kwargs...)

Applies a sddmSolver to the Laplacian of the adjacency matrix a of a connected graph. Passes on kwargs to the solver. sddmSolver should be a solver that obeys the interface.

Laplacians.lapWrapSDDM — Function.

f = lapWrapSDDM(sddmSolver, A::AbstractArray; tol::Real=1e-6, maxits=Inf, maxtime=Inf, verbose=false, pcgIts=Int[], params...)
f = lapWrapSDDM(sddmSolver)

Uses a sddmSolver to solve systems of linear equations in Laplacian matrices.

Laplacians.latin_square — Function.

S = latin_square(n)

Computes a random (but not uniformly random) n-by-n latin square.

Laplacians.latin_square_graph — Function.

a = latin_square_graph(S::Matrix{Int})
a = latin_square_graph(n::Int)

Construct the adjacency matrix of the latin square graph for the latin square S. If only n is provided, construct a latin square graph of dimension n.

Laplacians.ldli2Chol — Function.

L = ldli2Chol(ldli)

This produces a matrix L so that L L^T approximate the original Laplacians. It is not quite a Cholesky factor, because it is off by a perm (and the all-1s vector orthogonality.

Laplacians.llsAdd

Laplacians.llsInit

Laplacians.llsPurge

Laplacians.localBlockFlow — Function.

Compute block flow between s and t

Laplacians.localFlow — Function.

The LocalFlow function, from the Orecchia-Zhu paper

Laplacians.localImprove — Function.

localImprove{Tv,Ti}(G::SparseMatrixCSC{Tv,Ti}, A::Array{Int64,1}; epsSigma=-1.0, err=1e-10, maxSize = max(G.n, G.m)

The LocalImprove function, from the Orrechia-Zhu paper. Given a graph and an initial set, finds a set of smaller conductance based on the starting set using a localized version of max-flow.

Small discussion: When adding in the neighbors of the initial component, if the resulting conductance is worse than the initial one, the algorithm will add more and more vertices until hitting a better conductance. However, if we fix a certain maximum size for our component, it might be the case that this new conductance will always be worse than what we had initially. Thus, if we run the algorithm with a small maxSize, our initial conductance might be the best solution we can raech.

G is the given graph, A is the initial set

epsSigma is a measure of the quality of the returning set (the smaller the better). It's defaulted to volume(A) / volume(V - A)

err is the numerical error considered throughout the algorithm. It's defaulted to 1e-10

maxSize is the maximum allowed size for the flow graph at any iteration of the algorithm. It's defaulted to |V|

Laplacians.mapweight — Function.

b = mapweight(a, x->rand())

Create a new graph that is the same as the original, but with f applied to each nonzero entry of a. For example, to make the weight of every edge uniform in [0,1], we could write.

Laplacians.matToTree

Laplacians.matToTreeDepth

Laplacians.match_walk

Laplacians.maxflow — Function.

implementation of Dinic's algorithm. computes the maximum flow and min-cut in G between s and t we consider the adjacency matrix to be the capacity matrix

Laplacians.moduloNext

Laplacians.nbri

Laplacians.nbrs

Laplacians.nullSolver

Laplacians.pathFromParents

Laplacians.pathGraph

Laplacians.path_graph — Function.

graph = path_graph(n)

Laplacians.path_graph_ijv — Function.

ijv = path_graph_ijv(n::Int64)

Laplacians.pcg — Function.

x = pcg(mat, b, pre; tol, maxits, maxtime, verbose, pcgIts, stag_test)`

solves a symmetric linear system using preconditioner pre.

Arguments

pre can be a function or a matrix. If a matrix, a function to solve it is created with cholFact.

tol is set to 1e-6 by default,

maxits defaults to Inf

maxtime defaults to Inf. It measures seconds.

verbose defaults to false

pcgIts is an array for returning the number of pcgIterations. Default is length 0, in which case nothing is returned.

stag_test=k stops the code if rho[it] > (1-1/k) rho[it-k]. Set to 0 to deactivate.

Laplacians.pcgBLAS

Laplacians.pcgLapSolver — Function.

x = pcgLapSolver(A, B; tol::Real=1e-6, maxits=Inf, maxtime=Inf, verbose=false, pcgIts=Int[])

Create a solver that uses pcg to solve Laplacian systems in A Specialized for the case when the preconditioner the Laplacian matrix of B. It solves the preconditioner by Cholesky Factorization.

Laplacians.pcgSlow

Laplacians.pcgSolver — Function.

x = pcgSolver(mat, pre; tol, maxits, maxtime, verbose, pcgIts)

creates a solver for a PSD system using preconditioner pre. The parameters are as described in pcg.

Laplacians.plotGraph

Laplacians.plot_graph — Function.

plot_graph(gr,x,y;color=[0,0,1],dots=true,setaxis=true,number=false)

Plots graph gr with coordinates (x,y)

plot_graph(gr,x,y,z;color=[0,0,1],dots=true,setaxis=true,number=false)

Plots graph gr with coordinates (x,y,z)

Laplacians.power — Function.

ap = power(a::SparseMatrixCSC, k::Int)

Returns the kth power of a.

Laplacians.powerIteration

Laplacians.prefAttach

Laplacians.pref_attach — Function.

graph = pref_attach(n::Int64, k::Int64, p::Float64; ver=Vcur)

A preferential attachment graph in which each vertex has k edges to those that come before. These are chosen with probability p to be from a random vertex, and with probability 1-p to come from the endpoint of a random edge. It begins with a k-clique on the first k+1 vertices.

Laplacians.pref_attach_ijv

Laplacians.prim — Function.

prim(mat::SparseMatrixCSC; kind=:max) Compute a maximum spanning tree of the matrix mat. If kind=:min, computes a minimum spanning tree.

Laplacians.print_ll_col — Function.

Print a column in an LLmatp matrix. This is here for diagnostics.

Print a column in an LLMatOrd matrix. This is here for diagnostics.

Laplacians.printijv

Laplacians.prn — Function.

prn{Tv, Ti}(G::SparseMatrixCSC{Tv,Ti}, s::Array{Int64,1}, phi::Float64, b::Int64)

The PageRank-Nibble cutting algorithm from the Anderson/Chung/Lang paper

s is a set of starting vertices, phi is a constant in (0, 1], and b is an integer in [1, [log m]]

phi is a bound on the quality of the conductance of the cut - the smaller the phi, the higher the quality. b is used to handle precision throughout the algorithm - the higher the b, the greater the precision.

Laplacians.productGraph

Laplacians.product_graph — Function.

aprod = productGraph(a0, a1)

The Cartesian product of two graphs. When applied to two paths, it gives a grid.

Laplacians.pull!

Laplacians.pure_random_graph — Function.

graph = pure_random_graph(n::Integer; verbose=false, ver=Vcur)

Generate a random graph with n vertices from one of our natural distributions

Laplacians.pure_random_ijv — Function.

a = pure_random_ijv(n::Integer; verbose=false, prefix="", ver=Vcur)

Chooses among pathgraph, ringgraph, gridgraph, completebinarytree, randgenring, growngraph and ErdosRenyiClusterFix. It can produce a disconnected graph. For code that always produces a connected graph (and is the same as with Julia v0.6, use purerandomijv_v6)

Laplacians.pure_random_ijv_v6

Laplacians.pure_random_ijv_v7

Laplacians.pushSpeedResult! — Function.

ret is the answer returned by a speed test. This pushed it into the dictionary on which we are storing the tests.

Laplacians.randGenRing

Laplacians.randMatching

Laplacians.randRegular

Laplacians.randSet

Laplacians.randWeight

Laplacians.randWeightSub

Laplacians.rand_gen_ring — Function.

graph = rand_gen_ring(n, k; verbose = false, ver=Vcur)

A random generalized ring graph of degree k. Gens always contains 1, and the other k-1 edge types are chosen from an exponential distribution

Laplacians.rand_gen_ring_ijv

Laplacians.rand_matching — Function.

graph = rand_matching(n::Integer; ver=Vcur)

A random matching on n vertices

Laplacians.rand_matching_ijv

Laplacians.rand_regular — Function.

graph = rand_regular(n, k; ver=Vcur)

A sum of k random matchings on n vertices

Laplacians.rand_regular_bipartite — Function.

a = rand_regular_bipartite(n,k)

Random k-regular bipartite graph between two sets of n vertices. No repeat edges, so can take a long time to build of k is close to n.

Returns a (possibly) asymmetric matrix that contains the upper-right block.

Laplacians.rand_regular_ijv

Laplacians.rand_weight — Function.

graph = randWeight(graph; ver=Vcur)

Applies one of a number of random weighting schemes to the edges of the graph

Laplacians.rand_weight_sub

Laplacians.randishKruskal — Function.

tree = randishKruskal(A)

A heuristic for computing low-stretch spanning trees. Where Kruskal's MST algorithm adds edges in order of weight, this algorithm adds them at random with probability proportional to their weight.

Laplacians.randishPrim — Function.

tree = randishPrim(A)

A heuristic for computing low-stretch spanning trees. Where Prim's MST algorithm grows a cluster by always adding the edge on the boundary of maximum weight, this algorithm adds a boundary edge with probability proportional to its weight.

Random.randperm — Function.

graph = randperm(mat::AbstractMatrix)
        randperm(f::Expr)

Randomly permutes the vertex indices

Laplacians.randperm_ver!

Laplacians.readIJ — Function.

To read a simple edge list, each line being an (i, j) pair

Laplacians.readIJV — Function.

To read a simple edge list, each line being an (i, j, v) pair. The parens should not be there in the format, just commas separating. To generate this format in Matlab, you just need to be careful to write the vertex indices with sufficient precision. For example, you can do this

>> [ai,aj,av] = find(triu(a));
>> dlmwrite('graph.txt',[ai,aj,av],'precision',9);

Laplacians.refineCut — Function.

Modify a cluster by adding or removing vertices by picking at each step 
the vertex that has the maximum value of (Deg_external - Deg_Internal).
Each vertex can be added in/removed only once.

Laplacians.regular_bipartite_matching

Laplacians.remove!

Laplacians.reset!

Laplacians.reusableIntMap

Laplacians.ringGraph

Laplacians.ring_graph — Function.

graph = ring_graph(n)

Laplacians.ring_graph_ijv — Function.

ijv = ring_graph_ijv(n)

Laplacians.sample

Laplacians.sampleByWeight — Function.

ind = sampleByWeight(wt; ver=Vcur)

sample an index with probability proportional to its weight given here

Laplacians.sampleMany

Laplacians.sddmWrapLap — Function.

f = sddmWrapLap(lapSolver, sddm::AbstractArray; tol::Real=1e-6, maxits=Inf, maxtime=Inf, verbose=false, pcgIts=Int[], params...)

Uses a lapSolver to solve systems of linear equations in sddm matrices.

Laplacians.semiWtedChimera

Laplacians.semiwted_chimera — Function.

graph = semiwted_chimera(n::Integer; verbose=false, ver=Vcur)

A Chimera graph with some weights. The weights just appear when graphs are combined. For more interesting weights, use wted_chimera

Laplacians.semiwted_chimera_ijv

Laplacians.semiwted_chimera_ijv_v6

Laplacians.semiwted_chimera_ijv_v7

Laplacians.set!

Laplacians.setValue — Function.

Sets the value of a certain edge in a sparse graph; value can be 0 without the edges dissapearing

Laplacians.shortIntGraph — Function.

graph = shortIntGraph(a::SparseMatrixCSC)

Convert the indices in a graph to 32-bit ints. This takes less storage, but does not speed up much.

Laplacians.shortestPathTree — Function.

Computes the shortest path tree, and returns it as a sparse matrix. Treats edge weights as reciprocals of lengths. For example:

a = [0 2 1; 2 0 3; 1 3 0]
tr = full(shortestPathTree(sparse(a),1))

3x3 Array{Float64,2}:
 0.0  2.0  0.0
 2.0  0.0  3.0
 0.0  3.0  0.0

Laplacians.shortestPaths — Function.

Computes the lenghts of shortest paths from start. Returns both a vector of the lenghts, and the parent array in the shortest path tree.

This algorithm treats edge weights as reciprocals of distances. DOC BETTER

Laplacians.sortIJVind

Laplacians.sortSet — Function.

Given a set of integers, set between 1 and n, return a sorted version of them

Laplacians.sparsify — Function.

as = sparsify(a; ep=0.5)

Apply Spielman-Srivastava sparsification: sampling by effective resistances. ep should be less than 1.

Laplacians.spectralCoords

Laplacians.spectralDrawing

Laplacians.spectral_coords — Function.

spectral_coords(a)

Computes the spectral coordinates of a graph

Laplacians.spectral_drawing — Function.

spectral_drawing(a)

Computes spectral coordinates, and then uses plotGraph to draw

Laplacians.speedTestLapSolvers — Function.

function speedTestLapSolvers{Tv,Ti}(solvers, dic, a::SparseMatrixCSC{Tv,Ti}, b::Array{Tv,1}; tol::Real=1e-2, maxits=Inf, maxtime=Inf, verbose=false)

Runs many Laplacians solvers. Puts the build and solve time results into a dictionary dic. It would be easiest to look at it via DataFrame(dic). Returns the answer from the last solver. solvers should be an array of SolverTest.

Laplacians.stretchSample

Laplacians.subMean!

Laplacians.submatrixCSC — Function.

For a the submatrix of a with the entries of ijv indexed by list. The list must be sorted. This is equivalent to

(ai,aj,av) = findnz(a)
sparse(ai[list],aj[list],av[list],a.m,a.n)

Laplacians.subsampleEdges — Function.

graph = subsampleEdges(a::SparseMatrixCSC, p::Float64)

Create a new graph from the old, but keeping edge edge with probability p

Laplacians.support — Function.

sup12, sup21 = support(graph1, graph2; tol=1e-5)

Computes the support of graph1 wrt graph2, and the other way around. It is randomized, so you might want to run it again if you don't trust the answers.

Laplacians.symPermuteCSC — Function.

Permute the rows and columns of a symmetric matrix, a little faster than the default routine in Julia. It basically does a radix sort on rows than columns. It does require a to be symmetrix. But, the same ideas could probably speed up many sparse matrix routines. sympermute(a, perm) is the same as a[perm,perm].

Laplacians.symTransposeCSC — Function.

Compute the transpose of a matrix with symmetric nonzero structure

Laplacians.tarjanStretch

Laplacians.tarjanStretchSub

Laplacians.testId

Laplacians.testSolver

Laplacians.testZeroDiag — Function.

testZeroDiag(a)

Returns true if a has zero diagonal, false otherwise

Laplacians.test_latin_square

Laplacians.test_regular_bipartite

Laplacians.thicken — Function.

a_new = thicken(A,k)

Create a new graph with at least k times as many edges as A By connecting nodes with common neighbors at random. When this stops working (not enough new edges), repeat on the most recently produced graph. If k is too big, it is decreased so the average degree will not be pushed much above n/2.

When called without k, it just runs thicken_once.

For example:

a = grid2(5)
a2 = thicken(a,3)
(x,y) = grid2coords(5,5);
plotGraph(a2,x,y)

Laplacians.thickenOnce

Laplacians.thicken_once — Function.

a_new = thicken_once(a)

Creates one edge for every vertex in a of degree > 1 by connecting two of its random neighbors. To use this to thicken a, return unweight(a + a_new).

a = grid2(5)
a2 = unweight(a + thicken_once(a))
(x,y) = grid2coords(5,5);
plotGraph(a2,x,y)

Laplacians.toposort

Laplacians.treeDepthDFS — Function.

Compute the vector of depths in a tree that is in DFS order, with the root at the first position, and the leaves at the end

Laplacians.twoLift

Laplacians.two_lift — Function.

graph = two_lift(a, flip::AbstractArray{Bool,1})
graph = two_lift(a)
graph = two_lift(a, k::Integer)

Creats a 2-lift of a. flip is a boolean indicating which edges cross. In the third version, k is the number of edges that cross.

Laplacians.uniformWeight

Laplacians.uniformWeight! — Function.

uniformWeight!(a)

Set the weight of every edge to random uniform [0,1]

Laplacians.uniformWeight_ver — Function.

wted = uniformWeight(unwted)

Put a uniform [0,1] weight on every edge. This is an example of how to use mapweight.

Laplacians.uniformWeight_ver!

Laplacians.unweight — Function.

wt1 = unweight(a)

Create a new graph in that is the same as the original, but with all edge weights 1

Laplacians.unweight! — Function.

unweight!(a)

Change the weight of every edge in a to 1

Laplacians.vecToComps — Function.

This turns a component vector, like that generated by components, into an array of arrays of indices of vertices in each component. For example,

comps = vecToComps(c)

3-element Array{Array{Int64,1},1}:
 [1,2,3,4,6,7,8]
 [5,10]
 [9]

Laplacians.wdeg — Function.

Finds the weighted degree of a vertex in the graph

Laplacians.weighti

Laplacians.weights

Laplacians.wrapCapture — Function.

f = wrapCapture(solver::Function, mats, rhss; tol::Real=1e-6, maxits=Inf, maxtime=Inf, verbose=false, pcgIts=Int[], params...)

This wraps a solver so that we can capture all the matrices that it solves and all the right-hand-sides. Those are pushed into the arrays mats and rhss. For example

julia> mats = []
julia> rhss = []
julia> solver = wrapCapture(approxchol_lap, mats, rhss)
julia> a = chimera(10)
julia> f = solver(a);
julia> size(mats[1])
(10,10)
julia> b = randn(10)
julia> x = f(b);
julia> rhss
1-element Array{Any,1}:
 [0.404962,-0.827718,0.704616,-0.403223,0.204891,-0.505589,0.907015,1.90266,-0.438115,0.0464351]

Laplacians.wrapCaptureRhs — Function.

f = wrapCaptureRhs(sola::Function, rhss; tol::Real=1e-6, maxits=Inf, maxtime=Inf, verbose=false, pcgIts=Int[], params...)

Captures all the right-hand-sides that are passed to the solver sola. It pushes them into an array called rhhs. For example

julia> rhss = []
julia> a = wted_chimera(100)
julia> sola = approxchol_lap(a)
julia> wrappedSolver = wrapCaptureRhs(sola,rhss)
julia> b = randn(100)
julia> x = wrappedSolver(b,verbose=true)

PCG BLAS stopped after: 0.0 seconds and 11 iterations with relative error 3.160275810360986e-7.

julia> length(rhss[1])

100

Laplacians.wrapInterface — Function.

solveA = wrapInterface(solver::Function, A::AbstractMatrix; tol, maxits, maxtime, verbose, pcgIts=Int[],params...)
solverConstructor = wrapInterface(A::AbstractMatrix; tol, maxits, maxtime, verbose, pcgIts=Int[],params...)

Returns a function that discards tol, maxits, maxtime and verbose, sets pcgIts to 0 (because it might not be using pcg), and passes whatever params are left to the solver.

Examples

julia> a = randn(5,5);
julia> a = a * a';
julia> solvea = wrapInterface(X->cholesky(X,Val(true)), a, maxits=100, verbose=true);
julia> b = randn(5,1);
julia> norm(a*solvea(b, verbose=false)-b)
1.575705319704736e-14

julia> f = wrapInterface(X->cholesky(X,Val(true)))
julia> solvea = f(a, maxits=1000, maxtime = 1)
julia> norm(a*solvea(b, verbose=false, maxtime = 10)-b)
1.575705319704736e-14

Laplacians.writeIJV — Function.

Writes the upper portion of a matrix in ijv format, one row for each edge, separated by commas. Only writes the upper triangular portion. The result can be read from Matlab like this:

>> dl = dlmread('graph.txt');
>> a = sparse(dl(:,1),dl(:,2),dl(:,3));
>> n = max(size(a))
>> a(n,n) = 0;
>> a = a + a';

Laplacians.wtedChimera

Laplacians.wtedEdgeVertexMat — Function.

U = wtedEdgeVertexMat(a)

The signed and weighted edge-vertex adjacency matrix, so U'*U = L

Laplacians.wted_chimera — Function.

graph = wted_chimera(n::Integer, k::Integer; verbose=false, ver=Vcur)

Builds the kth wted chimeric graph on n vertices. It does this by resetting the random number generator seed. It should captute the state of the generator before that and then return it, but it does not yet.

graph = wted_chimera(n::Integer)

Generate a chimera, and then apply a random weighting scheme