Docstrings

WAV.KSDATAFORMAT_SUBTYPE_ALAW

WAV.KSDATAFORMAT_SUBTYPE_IEEE_FLOAT

WAV.KSDATAFORMAT_SUBTYPE_MULAW

WAV.KSDATAFORMAT_SUBTYPE_PCM

WAV.WAV

WAV.WAVArray

WAV.WAVChunk — Type.

A RIFF chunk.

WAV.WAVE_FORMAT_ALAW

WAV.WAVE_FORMAT_EXTENSIBLE

WAV.WAVE_FORMAT_IEEE_FLOAT

WAV.WAVE_FORMAT_MULAW

WAV.WAVE_FORMAT_PCM

WAV.WAVFormat

WAV.WAVFormatExtension

WAV.WAVMarker — Type.

A marker in a .wav file. start_time and duration are in samples.

WAV.WAVPlay

WAV.__init__

WAV.bits_per_sample

WAV.compress_sample_alaw

WAV.compress_sample_mulaw

WAV.convert_pcm_to_double

WAV.eval

WAV.format_length

WAV.get_default_compression

WAV.get_default_pcm_precision

WAV.get_default_precision

WAV.getformat

WAV.ieee_float_container_type

WAV.include

WAV.isextensible

WAV.isformat

WAV.load

WAV.make_range

WAV.pcm_container_type

WAV.read32

WAV.read_adtl

WAV.read_alaw_samples

WAV.read_companded_samples

WAV.read_cue — Function.

Read the contents of the cue chunk to extract marker start times

WAV.read_data

WAV.read_format

WAV.read_header

WAV.read_ieee_float_samples

WAV.read_le

WAV.read_list — Function.

Read the contents of the LIST chunk to extract marker names and durations

WAV.read_mulaw_samples

WAV.read_pcm_samples

WAV.read_tag

WAV.save

WAV.wav_cue_read — Function.

Return the markers from a list of chunks. Example:

using WAV
x, fs, bits, in_chunks = wavread("in.wav")
markers = wav_cue_read(in_chunks)

WAV.wav_cue_write — Function.

Write markers into a list of chunks. Example:

out_chunks = wav_cue_write(markers)
wavwrite(x, "out.wav", Fs=fs, nbits=16, compression=WAVE_FORMAT_PCM, chunks=out_chunks)

WAV.wav_info_read — Function.

Given a list of chunks as returned by wavread, return a Dict{Symbol, String} where the keys are symbols representing RIFF INFO tag IDs: https://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/RIFF.html#Info

WAV.wav_info_write — Function.

tags is a dictionary where the keys are symbols representing RIFF INFO tag IDs: https://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/RIFF.html#Info

Returns a list of chunks appropriate for passing to wavwrite.

WAV.wavappend

WAV.wavplay

WAV.wavread

WAV.wavwrite

WAV.write16

WAV.write32

WAV.write_companded_samples

WAV.write_cue

WAV.write_data

WAV.write_extended_header

WAV.write_format

WAV.write_header

WAV.write_ieee_float_samples

WAV.write_le

WAV.write_marker_list

WAV.write_pcm_samples

WAV.write_standard_header