Docstrings

REDCap.Config — Type.

REDCap.Config(url::String, key::String; ssl::Bool = true)

Struct to hold api url and key/superkey. APIConfigObj = Config("http...","ABCD...")

This will be passed to all function calls to both orient and authorize the api_pusher() function. The REDCap API cannot be accessed without this config object properly set-up. Always test your config object before automating a project.

NOTE: SSL should always be on by default.

If for whatever reason, your project will not SSL verify AND you must use it, disable ssl verification with ssl=false Leaving SSL verification disabled leaves you open for Man-in-the-Middle attacks and is generally just bad practice.

Parameters:

url - The url of the REDCap instance.

key - Either the standard or super API key.

ssl - Flag to enable ssl verification

REDCap.REDCap — Module.

REDCap

Legacy Julia 0.6 frontend for the REDCap API. Handles all available API calls from REDCap through top-level functions. Must have a valid REDCap.Config object set up with the API key and url. Compatable with REDCap Version 8.1.0

REDCap.api_pusher — Function.

api_pusher(mode::String, content::String, config::Config; format::String="", returnFormat::String="", file_loc::String="", kwargs...)

Pass the type of api call, the config struct, and any needed kwargs for that api call. Handles creation of the Dict of fields to pass to REDCap, and file IO/formatting.

API documentation found here: https://<your-redcap-site.com>/redcap/api/help/

Parameters:

mode - "import", "export", or "delete"

content - Passed by calling modules to indicate what data to access

config - Struct containing url and api-key

format - "json", "xml", "csv", or "odm". decides format of returned data

returnFormat - Error message format

file_loc - Location of file

kwargs... - Any addtl. arguments passed by the calling module

Returns:

Formatted response body

REDCap.create_project — Function.

create_project(config::REDCap.Config, project_title::String, purpose::Integer; format::String="json", returnFormat::String="json", odm="", purpose_other::String="", project_notes::String="", is_longitudinal::Integer=0, surveys_enabled::Integer=0, record_autonumbering_enabled::Integer=1)

Creates a project with the given parameters

Parameters:

config - Struct containing url and super-api-key

format - "json", "xml", "csv", or "odm". declares format of imported data

data - Attributes of project to create- only project_title and purpose are required (* for default)

* `project_title`: Title
* `purpose`: Must be numerical (0 - test, 1 - other, 2 - research, 3 - Qual+, 4 - OpSupport)
* `purpose_other`: If purpose == 1, string of purpose
* `project_notes`: Notes
* `is_longitudinal`: 0 - false*, 1 - true
* `surveys_enabled`: 0 - false*, 1 - true
* `record_autonumbering_enabled`: 0 - false, 1 - true*

returnFormat - Error message format

odm - XML string containing metadata

Returns:

The standard config for that project.

REDCap.delete_arms — Function.

delete_arms(config::REDCap.Config, arms::Array)

Delete Arms from project. Removing all arms reverts the project into a non-longitudinal project.

NOTE: This only works for longitudinal projects.

Parameters:

config - Struct containing url and api-key

arms - Array of arm names to delete

Returns:

Number of succesfully deleted arms

REDCap.delete_events — Function.

delete_events(config::REDCap.Config, events::Array)

Delete Events from project. Removing all but one event reverts the project into a non-longitudinal project.

NOTE: This only works for longitudinal projects.

Parameters:

config - Struct containing url and api-key

events - Array of event names to delete

Returns:

Number of successfully deleted events

REDCap.delete_file — Function.

delete_file(config::REDCap.Config, record::String, field::String, event::String; repeat_instance::Integer=1, returnFormat::String="json")

Delete document attached to record.

Parameters:

config - Struct containing url and api-key

record - Name of record containing file

field - Name of field containing file

event - Name of event containing file

repeat_instance - Number of repeated instances (long project)

returnFormat - Error message format

Returns:

Nothing/error

REDCap.delete_records — Function.

delete_records(config::REDCap.Config, records::Array; arm::Integer=0)

Delete one or more records from project.

Parameters:

config - Struct containing the url and api-key

records - Array of record names to delete

arm - Number of arm containing records

Returns:

Number of records successfully deleted

REDCap.df_formatter — Function.

df_formatter(data, mode::String)

Parameters:

data - The data to be formatted

mode - Formatting for Import (data to server) or Export (data from server)

Returns:

Either an JSON'ed dict, or a df

REDCap.df_parser — Function.

df_parser(data::Union{DataFrame, Dict})

Takes a DF, turns it into a Dict When a DF is passed, every row is turned into a dict() with the columns as keys, and pushed into an array to pass as a JSON object.

Parameters:

data - Data to be formatted

Returns:

A JSON ready dictionary array.

REDCap.eval

REDCap.export_arms — Function.

export_arms(config::REDCap.Config; arms::Array=[], format::String="json", returnFormat::String="json", file_loc::String="")

Returns a dict of all arms used in the project.

NOTE: This only works for longitudinal projects.

Parameters:

config - Struct containing url and api-key

arms - Array of arm names to export

format - "json", "xml", "csv", or "odm". decides format of returned data

returnFormat - Error message format

file_loc - Location to export to

Returns:

Formatted dict of Arms for project.

REDCap.export_events — Function.

export_events(config::REDCap.Config; arms::Array=[], format::String="json", returnFormat::String="json", file_loc::String="")

NOTE: This only works for longitudinal projects.

Parameters:

config - Struct containing url and api-key

arms - Array of arm names to export

format - "json", "xml", "csv", or "odm". decides format of returned data

returnFormat - Error message format

file_loc - Location to export to

Returns:

Formatted dict of events for project.

REDCap.export_field_names — Function.

export_field_names(config::REDCap.Config; field::String="", format::String="json", file_loc::String="")

Parameters:

config - Struct containing url and api-key

field - Specifies the field to export

format - "json", "xml", "csv", or "odm". decides format of returned data

file_loc - Location to export to

Returns:

Formatted dict of export/import-specific version of field names for all fields (or for one field, if desired) in project: 'originalfieldname', 'choicevalue', and 'exportfield_name'

REDCap.export_file — Function.

export_file(config::REDCap.Config, record::String, field::String, event::String; repeat_instance::Integer=1, returnFormat::String="json", file_loc::String="")

Parameters:

config - Struct containing url and api-key

record - Record id containing file

field - Field containing file

event - Event containing file

repeat_instance - Number of repeated instances (long. project)

returnFormat - Error message format

file_loc - Location to export to

Returns:

File attached to individual record.

REDCap.export_instrument_event_mappings — Function.

export_instrument_event_mappings(config::REDCap.Config, arms::Array=[]; format::String="json", returnFormat::String="json", file_loc::String="")

NOTE: This only works for longitudinal projects.

Parameters:

config - Struct containing url and api-key

arms - Array of arm names to export

format - "json", "xml", "csv", or "odm". decides format of returned data

returnFormat - Error message format

file_loc - Location to export to

Returns:

Formatted dict of instrument-event mappings for project.

REDCap.export_instruments — Function.

export_instruments(config::REDCap.Config; format::String="json", file_loc::String="")

Parameters:

config - Struct containing url and api-key

format - "json", "xml", "csv", or "odm". decides format of returned data

file_loc - Location to export to

Returns:

Formatted dict for data collection instruments of project.

REDCap.export_metadata — Function.

export_metadata(config::REDCap.Config; fields::Array=[], forms::Array=[], format::String="json", returnFormat::String="json", file_loc::String="")

Parameters:

config - Struct containing url and api-key

fields - Array of field names to pull data from

forms - Array of form names to pull data from

format - "json", "xml", "csv", or "odm". decides format of returned data

returnFormat - Error message format

file_loc - Location to export to

Returns:

Formatted dict of the metadata for project.

REDCap.export_pdf — Function.

export_pdf(config::REDCap.Config, file_loc::String; record::String="", event::String="", instrument::String="", allRecords::Bool=false)

Exports a PDF for a selected portion of the project.

Parameters:

config - Struct containing url and api-key

record - Record ID to populate PDF

event - Event name to populate PDF

instrument - Name of instrument to populate PDF

allRecords - Flag to take all records or not - if passed, the other specifying fields will be ignored

file_loc - Location to export to

Returns:

PDF file for:

single data collection instrument (blank),

all instruments (blank),

single instrument (with data from a single record),

all instruments (with data from a single record),

all instruments (with data from ALL records)

REDCap.export_project — Function.

export_project(config::REDCap.Config; returnMetadataOnly::Bool=false, records::Array=[], fields::Array=[], events::Array=[], format::String="xml", returnFormat::String="json", exportSurveyFields::Bool=false, exportDataAccessGroups::Bool=false, filterLogic::String="", exportFiles::Bool=false, file_loc::String="")

Parameters:

config - Struct containing url and api-key

returnMetadataOnly - Flag to return metedata or not

records - Array of record names to include

fields - Array of field names to include

events - Array of event names to include

returnFormat - Error message format

exportSurveyFields - Flag to return survey fields or not

exportDataAccessGroups - Flag to return DAGroups or not

filterLogic - Allows collection of records that fulfill a criteria eg. "[age] > 65"

exportFiles - Flag to include files or not

file_loc - Location to export to

Returns:

Entire project as XML.

REDCap.export_project_information — Function.

export_project_information(config::REDCap.Config; format::String="json", returnFormat::String="json", file_loc::String="")

Parameters:

config - Struct containing url and api-key

format - "json", "xml", "csv", or "odm". decides format of returned data

returnFormat - Error message format

file_loc - Location to export to

Returns:

Formatted dict of the basic attributes of given REDCap project.

REDCap.export_records — Function.

export_records(config::REDCap.Config; format::String="json", dtype::String="flat", records::Array=[], fields::Array=[], forms::Array=[], events::Array=[], rawOrLabel::String="raw", rawOrLabelHeaders::String="raw", exportCheckboxLabel::Bool=false, returnFormat::String="json", exportSurveyFields::Bool=false, exportDataAccessGroups::Bool=false, filterLogic::String="", file_loc::String="")

Parameters:

config - Struct containing url and api-key

format - "json", "xml", "csv", or "odm". decides format of returned data

dtype - Output mode: "flat" (output one record per row) or "eav" (one data point per row)

records - Array of record names to include

fields - Array of field names to include

forms - Array of form names to include

events - Array of event names to include

rawOrLabel - "raw" or "label" - export raw coded values or labels for multiple choice fields

rawOrLabelHeaders - Same as above, for headers

exportCheckboxLabel - Checkbox behavior: export checkboxes as "checked/unchecked" or as "field-name/'blank'"

returnFormat - Error message format

exportSurveyFields - Flag to return survey fields or not

exportDataAccessGroups - Flag to return DAGroups or not

filterLogic - Allows collection of records that fulfill a criteria eg. "[age] > 65"

file_loc - Location to export to

Returns:

An array of Dictionaries containing record information

REDCap.export_report — Function.

export_report(config::REDCap.Config, report_id::Integer; format::String="json", returnFormat::String="json", rawOrLabel::String="raw", rawOrLabelHeaders::String="raw", exportCheckboxLabel::Bool=false, file_loc::String="")

Parameters:

config - Struct containing url and api-key

report_id - Id of report to export

format - "json", "xml", "csv", or "odm". decides format of returned data

returnFormat - Error message format

rawOrLabel - "raw" or "label" - export raw coded values or labels for multiple choice fields

rawOrLabelHeaders - Same as above, for headers

exportCheckboxLabel - Checkbox behavior: export checkboxes as "checked/unchecked" or as "field-name/'blank'"

file_loc - Location to export to

Returns:

Formatted dict of report.

REDCap.export_survey_link — Function.

export_survey_link(config::REDCap.Config, record::String, instrument::String, event::String; repeat_instance::Int=1, returnFormat::String="json")

Parameters:

config - Struct containing url and api-key

record - Record id

instrument - Name of instrument linking to

event - Event name containing instrument

repeat_instance - Number of repeated instances (long project)

returnFormat - Error message format

Returns:

Unique survey link.

REDCap.export_survey_participant_list — Function.

export_survey_participant_list(config::REDCap.Config, instrument::String, event::String; format::String="json", returnFormat::String="json", file_loc::String="")

Parameters:

config - Struct containing url and api-key

instrument - Name of instrument to export list of participants

event - Event name conatining instrument

format - "json", "xml", "csv", or "odm". decides format of returned data

returnFormat - Error message format

file_loc - Location to export to

Returns:

Formatted dict of all participants for specific survey instrument.

REDCap.export_survey_queue_link — Function.

export_survey_queue_link(config::REDCap.Config, record::String; returnFormat::String="json")

Parameters:

config - Struct containing url and api-key

record - Record id for link

returnFormat - Error message format

Returns:

Unique Survey Queue link.

REDCap.export_survey_return_code — Function.

export_survey_return_code(config::REDCap.Config, record::String, instrument::String, event::String; repeat_instance::Integer=1, returnFormat::String="json")

Parameters:

config - Struct containing url and api-key

record - Record id for link

instrument - Name of instrument to export code for

event - event Name conatining instrument

repeat_instance - Number of repeated instances (long project)

returnFormat - Error message format

Returns:

Unique Return Code in plain text format.

REDCap.export_to_file — Function.

export_to_file(fileLoc::String, format::String, data)

Called by exporting functions to dump data into designated file, or yell at you for a bad path.

Parameters:

file_loc - Location of file - pass with proper extensions

data - The data to save to file

Returns:

Nothing/error

REDCap.export_users — Function.

export_users(config::REDCap.Config; format::String="json", returnFormat::String="json", file_loc::String="")

Parameters:

config - Struct containing url and api-key

format - "json", "xml", "csv", or "odm". decides format of returned data

returnFormat - Error message format

file_loc - Location to export to

Returns:

Array of formatted dicts of users for project.

REDCap.export_version — Function.

export_version(config::REDCap.Config; format::String="json")

Returns a string of the current REDCap version.

Parameters:

config - Struct containing url and api-key

format - "json", "xml", "csv", or "odm". decides format of returned data

Returns:

The version number (eg 1.0.0) as a string

REDCap.formatter — Function.

formatter(data, format, mode::String)

Takes data and sends out to the proper formating function.

Parameters:

data - The data to be formatted

format - The target format

mode - Formatting for Import (data to server) or Export (data from server)

Returns:

The specified formatted/unformatted object

REDCap.generate_next_record_id — Function.

generate_next_record_id(config::Config)

Parameters:

config - Struct containing url and api-key

Returns:

The next available ID number for project (Max record number +1)

REDCap.import_arms

REDCap.import_events — Function.

import_events(config::REDCap.Config, data; override::Int=0, format::String="json", returnFormat::String="json")

Update/import Events into a project.

Parameters:

config - Struct containing url and api-key

data - Data to be imported - pass as a file location to import from disk

override - 0 (false) 1 (true) - overwrites existing events

format - "json", "xml", "csv", or "odm". declares format of imported data

returnFormat - Error message format

Returns:

Number of successfully imported events

REDCap.import_file

REDCap.import_file_checker — Function.

import_file_checker()

Checks if the passed data is a valid path to a file, or data in itself. If a path, calls a loading function; if data, calls a formatter.

Parametes:

data - The data to check

format - The format to pass along

Returns:

The retreived/formatted data

REDCap.import_from_file — Function.

import_from_file(file_loc::String, format::String)

Called by importing functions to load already formatted data directly from a designated file

Parameters:

file_loc - Location of file

format - The target format

Returns:

The formatted data

REDCap.import_instrument_event_mappings

REDCap.import_metadata

REDCap.import_project_information — Function.

import_project_information(config::REDCap.Config, data; format::String="json")

Update basic attributes of given REDCap project. NOTE: Only for projects in development

Parameters:

config - Struct containing url and api-key

data - Data to be imported - pass as a file location to import from disk

format - "json", "xml", "csv", or "odm". declares format of imported data

Returns:

Number of successfully imported values

REDCap.import_records — Function.

import_records(config::REDCap.Config, data::Any; format::String="json", dtype::String="flat", overwriteBehavior::String="normal", forceAutoNumber::Bool=false, dateFormat::String="YMD", returnContent::String="count", returnFormat::String="json")

Import a set of records for a project.

Parameters:

config - Struct containing url and api-key

recordData - Array of record data to be imported - pass as a file location to import from disk

format - "json", "xml", "csv", or "odm". declares format of imported data

dtype - "flat" (one record per row) or "eav" (one data point per row)

overwriteBehavior - "normal" - will not overwrite, "overwrite" - will

forceAutoNumber - Force auto-numbering and overwrite given id number

dateFormat - "YMD", "MDY", or "DMY"

returnContent - "count" (number of successfully uploaded records),

"ids" (list of record numbers imported), 
					"auto-ids" (pair of assigned id and given id)

returnFormat - Error message format

Returns:

Specified by returnContent

REDCap.import_users

REDCap.include

REDCap.json_formatter — Function.

json_formatter(data, mode::String)

Parameters:

data - The data to be formatted

mode - Formatting for Import (data to server) or Export (data from server)

Returns:

Either a JSON'ed object or a parsed Dict

REDCap.odm_formatter

REDCap.poster — Function.

poster(config::Config, body)

Handles the POST duties for all modules. Also does basic Status checking and SSL verification.

Parameters:

config - Struct containing url and api-key

body - Request body data

Returns:

The response body.

REDCap.xml_formatter — Function.

xml_formatter(data, mode::String)

Parameters:

data - The data to be formatted

mode - Formatting for Import (data to server) or Export (data from server)

Returns:

Either an xml-formatted string, or an xml document