Docstrings

SpatialJackknife.SpatialJackknife

SpatialJackknife.eval

SpatialJackknife.get_dims — Function.

This function gets dimensions of a given data array and returns an error if they are not appropriate or consistent with the dimension of optional randoms.

SpatialJackknife.get_subvols — Function.

This function takes a dataset and returns an array of subvolume indices for each datapoint. In this case, it takes a set of random mask points for dividing arbitrary volumes. It computes the quantiles of the mask and then finds the closest mask point to each data point to determine the subvolume the data point belongs in.

The function expects a data array with shape (ndat, ndims) for Ndat samples in ndims dimensions. The mask array should have shape (nmask, ndims) and then there is an integere sidedivs which specifies the number of times each dimension is divided up to make sidedivs^ndims subvolumes.

As with the other method for get_subvols, this function computes the subvolume indices for given samples. It assumes a constant density throughout the volume. The method computes the subvolumes based on a set of extrema for sample values in each dimension and a number of volumes to divide the sample on per side. If a single set of extrema are given, the assumption is that they are the same in each dimension. Alternatively, it can take a set of volume edges in each dimension.

SpatialJackknife.include

SpatialJackknife.jackknife — Function.

This function computes the jackknife mean and variance over the observations computed from the data with obsfunc. They are computed by repeatedly applying obsfunc to the data with one of the subvolumes removed. It therefore assumes that obsfunc has normalised the observable for that remaining volume. The function obsfunc must be callable with the form

obsfunc(data::Array{Float64, 2}, args...)

for optional args tuple and the value returned by obsfunc must be in the form of an array of floats. If any of the returned values are NaNs, an error will be raised. By default, the covariance matrix is computed for observables in more than one dimension, but setting covar to false will result in only the diagonal variances being returned.