Docstrings

PEG.@rule — Macro.

@rule name = parsing_expression...

Defines a single rule in a parsing expression grammar as a function. See the documentation of the PEG module for a description of parsing expression grammars, and the type and usage of the rule functions defined using @rule.

PEG.Cache — Type.

A Cache maps from (rule name or gensym, input string length remaining) to either nothing or (parsed value, remaining input substring) (which is what rule functions return).

PEG.Failure — Type.

Singleton value that semantics functions may return to cause the parsing expression they were called from to fail, e.g. return Failure().

PEG.PEG — Module.

Define a Parsing Expression Grammar via a macro and abuse of Julia syntax.

Rules: @rule name = expression

Choice: infix ,

Sequence: infix &

Positive lookahead: prefix +

Negative lookahead: prefix -

Option (zero or one time): postfix [:?] (≡ [0:1])

(note that [?] won't work in Julia >= 1.0 per JuliaLang/julia#22712)

Zero or more times: postfix [*] (≡ [0:end])

One or more times: postfix [+] (≡ [1:end])

Exactly m times: postfix [m] (≡ [m:m]) (where m is an integer)

Between m and n times inclusive: postfix [m:n]

At most n times: postfix [0:n]

At least m times: postfix [m:end]

Terminals: r"regex", "string"

Extra regex flags: p is for punctuation, and eats whitespace (\s*) after the match; w is for word, and implies p, but also makes sure match boundaries are word boundaries (\b); h modifies p and w to eat only horizontal whitespace (\h). Values passed to semantics functions exclude eaten whitespace.

Semantics: expression |> unary_function (like ParserCombinator)

or expression > nary_function to interpolate args.

Returning the special singleton value PEG.Failure() from a semantics function causes the parsing expression it's attached to to fail (return nothing instead of a tuple). Returning nothing from a semantics function is not special; it just makes the first part of the tuple nothing. See the parsing function signature below.

Put another way:

using PEG
@rule grammar = "using PEG\n" & rule[*]
@rule rule = r"@rule"p & nonterminal & r"="p & choice
@rule choice = seq & (r","p & seq)[*]
@rule seq = item & (r"&"p & item)[*] & (r"\|?>"p & julia_function)[:?]
@rule item = lookahead , counted
@rule lookahead = r"\("p & (r"[+-]"p) & seq & r"\)"p
@rule counted = single & (count)[:?]
@rule count = range , r"\["p & (":?" , r"[\*\+]"p) & r"]"p
@rule range = r"\["p & integer & (r":"p & (integer , r"end"w))[:?] & r"]"p
@rule integer = r"\d+"w
@rule single = parens , terminal , nonterminal
@rule parens = r"\("p & choice & r"\)"p
@rule nonterminal = r"\pL\w+"w
@rule terminal = regex , string & r"\s*"
@rule regex = r"\br" & string & r"[himpswx]*\s*"
@rule string = r""(\\.|[^"])*""
@rule julia_function = # left as an exercise ;)

Each rule defines a parsing function with the following signature:

nonterminal(input::T, cache=PEG.Cache()) where T <: AbstractString
  ::Union{Nothing,Tuple{Any,SubString}}

The Any part of the return value is the abstract syntax tree, while the SubString is the remaining input after the parsed portion. If parsing fails, nothing is returned.

While you can use rules defined in this way directly, it might be more convenient to use the functions parse_next(rule, input; whole=false) or parse_whole(rule, input). See their documentation for more information.

Call PEG.setdebug!() to have debugging information printed during parsing. Call PEG.setdebug!(false) to turn it off again.

PEG.cache_rule

PEG.debug

PEG.eval

PEG.flatten_op — Function.

Given e.g. x=:(((a op b) op c) op d), return [a,b,c,d].

PEG.include

PEG.parse_next — Function.

parse_next(rule, input; whole=false)

Parse a prefix of the input string as one instance of the given rule. This differs from just calling the rule itself on the input in a few important ways:

When parsing succeeds, parse_next returns only the parsed value, while rule returns a Tuple of the parsed value and the remaining unparsed input.

When parsing fails, parse_next throws an exception, with information on what failed to match starting at the latest point in the string that any parsing expression matched up to. rule just returns nothing.

When whole=true, parsing only succeeds if the whole input is consumed, i.e. the second value in the tuple that rule returns is "".

PEG.parse_whole — Function.

parse_whole(rule, input) = parse_next(rule, input; whole=true)

PEG.remove_re_flag — Function.

remove_re_flag(flags, flag)

Return (a copy of) regex flags with flag removed, and a boolean indicating whether the flag was there in the first place.

PEG.setdebug! — Function.

setdebug!(val=true)

Enable or disable printing of debugging information.

PEG.squash_ast — Function.

squash_ast(x)

Given an ast returned by a grammar with no semantics, return a simplified version, such that there are no nothings, and no Any[]s with exactly zero or one element.

PEG.to_rule