Docstrings

LocalFilters.BilateralFilter

LocalFilters.CartesianRegion — Constant.

CartesianRegion{N} is an union of the types of anything suitable to define a Cartesian region in N dimensions. That is, an interval of Cartesian indices in N dimensions. Methods initialindex, finalindex, limits and cartesianregion can be applied to anything whose type belongs to CartesianRegion.

LocalFilters.Dimensions — Type.

Dimensions{N} is an N-tuple of integers, that is the type of an argument suitable for specifying a list of dimensions.

LocalFilters.IndexInterval — Constant.

IndexInterval is an union of the types of any argument suitable to specify an interval of indices along a dimension (that is, an integer or an integer valued unit range).

LocalFilters.Kernel — Type.

A Kernel can be used to define a weighted neighborhood (for weighted local average or for convolution) or a structuring element (for mathematical morphology). It is a rectangular array of coefficients over a, possibly off-centered, rectangular neighborhood. In general, it is sufficient to specify ::Kernel{T,N} in the signature of methods, with T the type of the coefficients and N the number of dimensions (the third parameter A of the type is to fully qualify the type of the array of coefficients).

A kernel is built as:

B = Kernel([T,] C, start=defaultstart(C))

where C is the array of coefficients (which can be retrieved by coefs(B)) and start the initial CartesianIndex for indexing the kernel (which can be retrieved by initialindex(B)). The start parameter let the caller choose an arbitrary origin for the kernel coefficients; when a filter is applied, the following mapping is assumed:

B[k] ≡ C[k + off]

where off = initialindex(C) - initialindex(B).

If start is omitted, its value is set so that the origin (whose index is zero(CartesianIndex{N}) with N the number of dimensions) of the kernel indices is at the geometric center of the array of coefficients (see LocalFilters.defaultstart). Optional argument T is to impose the type of the coefficients.

To convert the element type of the coefficients of an existing kernel, do:

Kernel(T, K)

which yields a kernel whose coefficients are those of the kernel K converted to type T.

It is also possible to convert instances of RectangularBox into a kernel with boolean coefficients by calling:

Kernel(B)

where B is the neighborhood to convert into an instance of Kernel.

LocalFilters.LocalFilters

LocalFilters.Neighborhood — Type.

All neighborhoods are instances of a type derived from Neighborhood.

LocalFilters.RectangularBox — Type.

RectangularBox(start, stop)

yields a neighborhood which is a rectangular (Cartesian) box defined by the bounds of the multi-dimensional indices in the box.

Another possibility is to specify the dimensions of the box and the offsets of its central element:

RectangularBox(dims, offs)

with dims a N-tuple of dimensions and offs either a N-tuple of indices of an instance of CartesianIndex{N}.

A RectangularBox can also be defined by the index ranges along all the dimensions. For example:

RectangularBox(-3:3, 0, -2:1)
RectangularBox((-3:3, 0, -2:1))

both yield a 3-dimensional RectangularBox of size 7×1×4 and whose first index varies on -3:3, its second index is 0 while its third index varies on -2:1.

Finally, a RectangularBox can be defined as:

RectangularBox(R)

where R is an instance of CartesianIndices or, on version of Julia ≤ 0.6, CartesianRange.

LocalFilters.Separable

LocalFilters.USE_CARTESIAN_RANGE

LocalFilters.UnitIndexRange — Type.

UnitIndexRange is any integer valued range with unit step, that is the type of an argument suitable for specifying a range of contiguous indices.

LocalFilters.UnitIndexRanges — Type.

UnitIndexRanges{N} is an N-tuple of UnitIndexRange, that is the type of an argument suitable for specifying a list of Cartesian indices.

LocalFilters._ball!

LocalFilters._length

LocalFilters._range

LocalFilters._store! — Function.

_store!(arr, idx, val) stores value val in array arr at index idx, taking care of rounding val if it is of floating-point type while the elements of arr are integers.

LocalFilters._typeofsum — Function.

_typeofsum(T) yields a numerical type suitable for storing a sum of elements of type T.

LocalFilters.ball — Function.

LocalFilters.ball(N, r)

yields a boolean mask which is a N-dimensional array with all dimensions odd and equal and set to true where position is inside a N-dimesional ball of radius r.

LocalFilters.BilateralFilter.bilateralfilter — Function.

bilateralfilter([T,] A, Fr, Gs, ...)

yields the result of applying the bilateral filter on array A.

Argument Fr specifies the range kernel for smoothing differences in intensities, it is a function which takes two values from A as arguments and returns a nonnegative value.

Arguments Gs, ... specify the spatial kernel for smoothing differences in coordinates.

Optional argument T can be used to force the floating-point type used for (most) computations.

The in-place version is:

bilateralfilter!([T,] dst, A, Fr, Gs, ...)

which stores in dst the result of applying the bilateral filter on array A.

See wikipedia.

LocalFilters.BilateralFilter.bilateralfilter! — Function.

bilateralfilter([T,] A, Fr, Gs, ...)

yields the result of applying the bilateral filter on array A.

Argument Fr specifies the range kernel for smoothing differences in intensities, it is a function which takes two values from A as arguments and returns a nonnegative value.

Arguments Gs, ... specify the spatial kernel for smoothing differences in coordinates.

Optional argument T can be used to force the floating-point type used for (most) computations.

The in-place version is:

bilateralfilter!([T,] dst, A, Fr, Gs, ...)

which stores in dst the result of applying the bilateral filter on array A.

See wikipedia.

LocalFilters.bottom_hat — Function.

top_hat(A, R)
top_hat(A, R, S)
bottom_hat(A, R)
bottom_hat(A, R, S)

Perform A summit/valley detection by applying a top-hat filter to array A. Argument R defines the structuring element for the feature detection. Optional argument S specifies the structuring element used to apply a smoothing to A prior to the top-hat filter. If R and S are specified as the radii of the structuring elements, then S should be smaller than R. For instance:

top_hat(bitmap, 3, 1)

may be used to detect text or lines in a bimap image.

The in-place versions:

top_hat!(dst, wrk, src, R)
bottom_hat!(dst, wrk, src, R)

apply the top-hat filter on the source src and store the result in the destination dst using wrk as a workspace array. These 3 arguments must be similar but different arrays. The destination dst is returned.

See also dilate, closing and morph_enhance.

LocalFilters.bottom_hat! — Function.

top_hat(A, R)
top_hat(A, R, S)
bottom_hat(A, R)
bottom_hat(A, R, S)

Perform A summit/valley detection by applying a top-hat filter to array A. Argument R defines the structuring element for the feature detection. Optional argument S specifies the structuring element used to apply a smoothing to A prior to the top-hat filter. If R and S are specified as the radii of the structuring elements, then S should be smaller than R. For instance:

top_hat(bitmap, 3, 1)

may be used to detect text or lines in a bimap image.

The in-place versions:

top_hat!(dst, wrk, src, R)
bottom_hat!(dst, wrk, src, R)

apply the top-hat filter on the source src and store the result in the destination dst using wrk as a workspace array. These 3 arguments must be similar but different arrays. The destination dst is returned.

See also dilate, closing and morph_enhance.

LocalFilters.cartesianregion — Function.

cartesianregion(args...) -> R

yields the rectangular region (as an instance of CartesianIndices or CartesianRange depending on Julia version) specified by the arguments which can be:

an abstract array whose axes define the region (see axes);

a list of unit range indices and/or indices along each dimension;

the corners of the bounding box, say start and stop, specified as two instances of CartesianIndex;

a neighborhood (see Neighborhood);

an instance of CartesianIndices or CartesianRange.

This method is a workaround to deal with optimization issues between different versions of Julia. In recent Julia versions (≥ 0.7), cartesianregion(args...) yields an instance of CartesianIndices; while in Julia version 0.6, cartesianregion(args...) yields a CartesianRange which appears to be faster than CartesianIndices as provided by Compat.

See also: initialindex, finalindex and limits.

LocalFilters.closing — Function.

closing(A, R)
opening(A, R)

respectively perform a closing or an opening of array A by the structuring element R. If not specified, R is a box of size 3 along all the dimensions of A. A closing is a dilation followed by an erosion, whereas an opening is an erosion followed by a dilation.

The in-place versions are:

closing!(dst, wrk, src, R)
opening!(dst, wrk, src, R)

which perform the operation on the source src and store the result in destination dst using wrk as a workspace array. These 3 arguments must be similar arrays, dst and src may be identical, but wrk must not be the same array as src or dst. The destination dst is returned.

See erode or dilate for the meaning of the arguments.

LocalFilters.closing! — Function.

closing(A, R)
opening(A, R)

respectively perform a closing or an opening of array A by the structuring element R. If not specified, R is a box of size 3 along all the dimensions of A. A closing is a dilation followed by an erosion, whereas an opening is an erosion followed by a dilation.

The in-place versions are:

closing!(dst, wrk, src, R)
opening!(dst, wrk, src, R)

which perform the operation on the source src and store the result in destination dst using wrk as a workspace array. These 3 arguments must be similar arrays, dst and src may be identical, but wrk must not be the same array as src or dst. The destination dst is returned.

See erode or dilate for the meaning of the arguments.

LocalFilters.coefs — Function.

LocalFilters.coefs(B) yields the array of coefficients embedded in kernel B.

See also: LocalFilters.offset.

LocalFilters.convertcoefs — Function.

convertcoefs(T, A)

yields an array of kernel coefficients equivalent to array A but whose elements have type T.

If T is a floating-point type and A is a boolean array, then the values of the result are one(T) where A is true and zero(T) elsewhere. To use different values (for instance, to define flat structuring elements), you may call:

convertcoefs((vtrue, vfalse), A)

with A a boolean array to get an array whose elements are equal to vtrue where A is true and to vfalse otherwise.

See also: strel.

LocalFilters.convolve — Function.

convolve(A, B)

yields the convolution of A by the support of the neighborhood defined by B of by the kernel B if it is an instance of LocalFilters.Kernel with numerical coefficients. The result is an array similar to A.

The in-place version is:

convolve!(dst, A, B) -> dst

See also localfilter!.

LocalFilters.convolve! — Function.

convolve(A, B)

yields the convolution of A by the support of the neighborhood defined by B of by the kernel B if it is an instance of LocalFilters.Kernel with numerical coefficients. The result is an array similar to A.

The in-place version is:

convolve!(dst, A, B) -> dst

See also localfilter!.

LocalFilters.defaultstart — Function.

defaultstart(A) -> I::CartesianIndex

yields the initial (multi-dimensional) index of a rectangular region which has the same size as the array A but whose origin (that is, index zero(CartesianIndex{N})) is at the geometrical center of the region (with the same conventions as fftshift).

LocalFilters.dilate — Function.

Basic operations of mathematical morphology are:

erode(A, B) -> Amin
dilate(A, B) -> Amax

which respectively return the local minima Amin and the local maxima Amax of argument A in a neighborhood defined by B. The returned result is similar to A (same size and type).

The two operations can be combined in one call:

localextrema(A, B) -> Amin, Amax

The in-place versions:

erode!(Amin, A, B) -> Amin
dilate!(Amax, A, B) -> Amax
localextrema!(Amin, Amax, A, B) -> Amin, Amax

apply the operation to A with structuring element B and store the result in the provided arrays Amin and/or Amax.

See also localmean, opening, closing, top_hat and bottom_hat.

LocalFilters.dilate! — Function.

Basic operations of mathematical morphology are:

erode(A, B) -> Amin
dilate(A, B) -> Amax

which respectively return the local minima Amin and the local maxima Amax of argument A in a neighborhood defined by B. The returned result is similar to A (same size and type).

The two operations can be combined in one call:

localextrema(A, B) -> Amin, Amax

The in-place versions:

erode!(Amin, A, B) -> Amin
dilate!(Amax, A, B) -> Amax
localextrema!(Amin, Amax, A, B) -> Amin, Amax

apply the operation to A with structuring element B and store the result in the provided arrays Amin and/or Amax.

See also localmean, opening, closing, top_hat and bottom_hat.

LocalFilters.erode — Function.

Basic operations of mathematical morphology are:

erode(A, B) -> Amin
dilate(A, B) -> Amax

which respectively return the local minima Amin and the local maxima Amax of argument A in a neighborhood defined by B. The returned result is similar to A (same size and type).

The two operations can be combined in one call:

localextrema(A, B) -> Amin, Amax

The in-place versions:

erode!(Amin, A, B) -> Amin
dilate!(Amax, A, B) -> Amax
localextrema!(Amin, Amax, A, B) -> Amin, Amax

apply the operation to A with structuring element B and store the result in the provided arrays Amin and/or Amax.

See also localmean, opening, closing, top_hat and bottom_hat.

LocalFilters.erode! — Function.

Basic operations of mathematical morphology are:

erode(A, B) -> Amin
dilate(A, B) -> Amax

which respectively return the local minima Amin and the local maxima Amax of argument A in a neighborhood defined by B. The returned result is similar to A (same size and type).

The two operations can be combined in one call:

localextrema(A, B) -> Amin, Amax

The in-place versions:

erode!(Amin, A, B) -> Amin
dilate!(Amax, A, B) -> Amax
localextrema!(Amin, Amax, A, B) -> Amin, Amax

apply the operation to A with structuring element B and store the result in the provided arrays Amin and/or Amax.

See also localmean, opening, closing, top_hat and bottom_hat.

LocalFilters.eval

LocalFilters.finalindex

LocalFilters.include

LocalFilters.initialindex — Function.

initialindex(B) -> Imin::CartesianIndex{N}
finalindex(B)   -> Imax::CartesianIndex{N}

respectively yield the initial and final multi-dimensional index for indexing the Cartesian region defined by B. A Cartesian region defines a rectangular set of indices whose edges are aligned with the indexing axes.

Compared to similar methods firstindex, lastindex(), first() and last(), the returned value is always an instance of CartesianIndex{N} with N the number of dimensions.

Any multi-dimensional index I::CartesianIndex{N} is in the Cartesian region defined B if and only if Imin ≤ I ≤ Imax.

Also see: limits.

LocalFilters.ismmbox — Function.

ismmbox(B)

yields whether neighborhood B has the same effect as a rectangular box for mathematical morphology operations. This may be used to use fast separable versions of mathematical morphology operations like the van Herk-Gil-Werman algorithm.

See also: LocalFilters.RectangularBox.

LocalFilters.limits — Function.

limits(T::DataType) -> typemin(T), typemax(T)

yields the infimum and supremum of a type T.

limits(B) -> Imin, Imax

yields the corners (as a tuple of 2 CartesianIndex) of the Cartesian region defined by B.

Also see: initialindex and finalindex.

LocalFilters.localextrema — Function.

Basic operations of mathematical morphology are:

erode(A, B) -> Amin
dilate(A, B) -> Amax

which respectively return the local minima Amin and the local maxima Amax of argument A in a neighborhood defined by B. The returned result is similar to A (same size and type).

The two operations can be combined in one call:

localextrema(A, B) -> Amin, Amax

The in-place versions:

erode!(Amin, A, B) -> Amin
dilate!(Amax, A, B) -> Amax
localextrema!(Amin, Amax, A, B) -> Amin, Amax

apply the operation to A with structuring element B and store the result in the provided arrays Amin and/or Amax.

See also localmean, opening, closing, top_hat and bottom_hat.

LocalFilters.localextrema! — Function.

Basic operations of mathematical morphology are:

erode(A, B) -> Amin
dilate(A, B) -> Amax

which respectively return the local minima Amin and the local maxima Amax of argument A in a neighborhood defined by B. The returned result is similar to A (same size and type).

The two operations can be combined in one call:

localextrema(A, B) -> Amin, Amax

The in-place versions:

erode!(Amin, A, B) -> Amin
dilate!(Amax, A, B) -> Amax
localextrema!(Amin, Amax, A, B) -> Amin, Amax

apply the operation to A with structuring element B and store the result in the provided arrays Amin and/or Amax.

See also localmean, opening, closing, top_hat and bottom_hat.

LocalFilters.localfilter — Function.

General local filters

A local filtering operation can be performed by calling:

localfilter!(dst, A, B, initial, update, store) -> dst

where dst is the destination, A is the source, B defines the neighborhood, initial, update and store are three functions whose purposes are explained by the following pseudo-code to implement the local operation:

for i ∈ Sup(A)
    v = initial(A[i])
    for j ∈ Sup(A) and i - j ∈ Sup(B)
        v = update(v, A[j], B[i-j])
    end
    store(dst, i, v)
end

where A Sup(A) yields the support of A (that is the set of indices in A) and likely Sup(B) for B.

For instance, to compute a local minimum (that is, an erosion):

localfilter!(dst, A, B,
             (a)     -> typemax(T),
             (v,a,b) -> min(v,a),
             (d,i,v) -> d[i] = v)

Important: Because the result of an elementary operation can be something else than just a scalar, the loop(s) in localfilter! are performed without bound checking of the destination and it is the caller's responsability to insure that the destination have the correct sizes.

Local filter by the van Herk-Gil-Werman algorithm

localfilter!([dst = A,] A, dims, op, rngs [, w])

overwrite the contents of dst with the result of applying van Herk-Gil-Werman algorithm to filter array A along dimension(s) dims with (associative) binary operation op and contiguous structuring element(s) defined by the interval(s) rngs. Optional argument w is a workspace array which is automatically allocated if not provided; otherwise, it must be a vector of same element type as A which is resized (with resize!) as needed. The destination dst must have the same indices as the source A (see axes). Operation can be done in-place; that is, dst and A can be the same.

Argument dims specifies along which dimension(s) of A the filter is to be applied, it can be a single integer, several integers or a colon : to specify all dimensions. Dimensions are processed in the order given by dims (the same dimension may appear several times) and there must be a matching interval in rngs to specify the structuring element (except that if rngs is a single interval, it is used for every dimension in dims). An interval is either an integer or an integer unit range in the form kmin:kmax (an interval specified as a single integer, say k, is the same as specifying k:k).

Assuming mono-dimensional arrays A and dst, the single filtering pass:

localfilter!(dst, A, :, op, rng)

yields:

dst[j] = A[j-kmax] ⋄ A[j-kmax+1] ⋄ A[j-kmax+2] ⋄ ... ⋄ A[j-kmin]

for all j ∈ [first(axes(A,1)):last(axes(A,1))], with x ⋄ y = op(x, y), kmin = first(rng) and kmax = last(rng). Note that if kmin = kmax = k (which occurs if rng is a simple integer), the result of the filter is to operate a simple shift by k along the corresponding dimension and has no effects if k = 0. This can be exploited to not filter some dimension(s).

The out-place version, allocates the destination array and is called as:

localfilter([T,] A, dims, op, rngs [, w])

with T the element type of the result (by default T = eltype(A)).

Examples

The in-place morphological erosion (local minimum) of the array A on a centered structuring element of width 7 in every dimension can be applied by:

localfilter!(A, :, min, -3:3)

One can specify index interval 0 to do nothing along the corresponding dimension. For instance:

localfilter!(A, :, max, (-3:3, 0, -4:4))

will overwrite A with the local maxima (a.k.a. morphological dilation) of the three-dimensional array A in a centered local neighborhood of size 7×1×9 (nothing is done along the second dimension). The same result may be obtained with:

localfilter!(A, (1,3), max, (-3:3, -4:4))

where the second dimension is omitted from the list of dimensions.

The local average of the two-dimensional array A on a centered structuring element of size 11×11 can be computed as:

localfilter(A, :, +, (-5:5, -5:5))*(1/11)

Efficiency and restrictions

The van Herk-Gil-Werman algorithm is very fast for rectangular structuring elements. It takes at most 3 operations to filter an element along a given dimension whatever the width p of the considered local neighborhood. For N-dimensional arrays, the algorithm requires only 3N operations per element instead of p^N - 1 operations for a naive implementation. This however requires to make a pass along each dimension so memory page faults may reduce the performances. This is somewhat attenuated by the fact that the algorithm can be applied in-place. For efficient multi-dimensional out-of-place filtering, it is recommended to make the first pass with a fresh destination array and then all other passes in-place on the destination array.

To apply the van Herk-Gil-Werman algorithm, the structuring element must be separable along the dimensions and its components must be contiguous. In other words, the algorithm is only applicable for N-dimensional rectangular neighborhoods. The structuring element may however be off-centered by arbitrary offsets along each dimension.

To take into account boundary conditions (for now only least neighborhood is implemented) and allow for in-place operation, the algorithm allocates a workspace array.

LocalFilters.localfilter! — Function.

Local filter by the van Herk-Gil-Werman algorithm

localfilter!([dst = A,] A, dims, op, rngs [, w])

overwrite the contents of dst with the result of applying van Herk-Gil-Werman algorithm to filter array A along dimension(s) dims with (associative) binary operation op and contiguous structuring element(s) defined by the interval(s) rngs. Optional argument w is a workspace array which is automatically allocated if not provided; otherwise, it must be a vector of same element type as A which is resized (with resize!) as needed. The destination dst must have the same indices as the source A (see axes). Operation can be done in-place; that is, dst and A can be the same.

Argument dims specifies along which dimension(s) of A the filter is to be applied, it can be a single integer, several integers or a colon : to specify all dimensions. Dimensions are processed in the order given by dims (the same dimension may appear several times) and there must be a matching interval in rngs to specify the structuring element (except that if rngs is a single interval, it is used for every dimension in dims). An interval is either an integer or an integer unit range in the form kmin:kmax (an interval specified as a single integer, say k, is the same as specifying k:k).

Assuming mono-dimensional arrays A and dst, the single filtering pass:

localfilter!(dst, A, :, op, rng)

yields:

dst[j] = A[j-kmax] ⋄ A[j-kmax+1] ⋄ A[j-kmax+2] ⋄ ... ⋄ A[j-kmin]

for all j ∈ [first(axes(A,1)):last(axes(A,1))], with x ⋄ y = op(x, y), kmin = first(rng) and kmax = last(rng). Note that if kmin = kmax = k (which occurs if rng is a simple integer), the result of the filter is to operate a simple shift by k along the corresponding dimension and has no effects if k = 0. This can be exploited to not filter some dimension(s).

The out-place version, allocates the destination array and is called as:

localfilter([T,] A, dims, op, rngs [, w])

with T the element type of the result (by default T = eltype(A)).

Examples

The in-place morphological erosion (local minimum) of the array A on a centered structuring element of width 7 in every dimension can be applied by:

localfilter!(A, :, min, -3:3)

One can specify index interval 0 to do nothing along the corresponding dimension. For instance:

localfilter!(A, :, max, (-3:3, 0, -4:4))

will overwrite A with the local maxima (a.k.a. morphological dilation) of the three-dimensional array A in a centered local neighborhood of size 7×1×9 (nothing is done along the second dimension). The same result may be obtained with:

localfilter!(A, (1,3), max, (-3:3, -4:4))

where the second dimension is omitted from the list of dimensions.

The local average of the two-dimensional array A on a centered structuring element of size 11×11 can be computed as:

localfilter(A, :, +, (-5:5, -5:5))*(1/11)

Efficiency and restrictions

The van Herk-Gil-Werman algorithm is very fast for rectangular structuring elements. It takes at most 3 operations to filter an element along a given dimension whatever the width p of the considered local neighborhood. For N-dimensional arrays, the algorithm requires only 3N operations per element instead of p^N - 1 operations for a naive implementation. This however requires to make a pass along each dimension so memory page faults may reduce the performances. This is somewhat attenuated by the fact that the algorithm can be applied in-place. For efficient multi-dimensional out-of-place filtering, it is recommended to make the first pass with a fresh destination array and then all other passes in-place on the destination array.

To apply the van Herk-Gil-Werman algorithm, the structuring element must be separable along the dimensions and its components must be contiguous. In other words, the algorithm is only applicable for N-dimensional rectangular neighborhoods. The structuring element may however be off-centered by arbitrary offsets along each dimension.

To take into account boundary conditions (for now only least neighborhood is implemented) and allow for in-place operation, the algorithm allocates a workspace array.

General local filters

A local filtering operation can be performed by calling:

localfilter!(dst, A, B, initial, update, store) -> dst

where dst is the destination, A is the source, B defines the neighborhood, initial, update and store are three functions whose purposes are explained by the following pseudo-code to implement the local operation:

for i ∈ Sup(A)
    v = initial(A[i])
    for j ∈ Sup(A) and i - j ∈ Sup(B)
        v = update(v, A[j], B[i-j])
    end
    store(dst, i, v)
end

where A Sup(A) yields the support of A (that is the set of indices in A) and likely Sup(B) for B.

For instance, to compute a local minimum (that is, an erosion):

localfilter!(dst, A, B,
             (a)     -> typemax(T),
             (v,a,b) -> min(v,a),
             (d,i,v) -> d[i] = v)

Important: Because the result of an elementary operation can be something else than just a scalar, the loop(s) in localfilter! are performed without bound checking of the destination and it is the caller's responsability to insure that the destination have the correct sizes.

LocalFilters.localmean — Function.

localmean(A, B)

yields the local mean of A in a neighborhood defined by B. The result is an array similar to A.

The in-place version is:

localmean!(dst, A, B) -> dst

See also localfilter!.

LocalFilters.localmean! — Function.

localmean(A, B)

yields the local mean of A in a neighborhood defined by B. The result is an array similar to A.

The in-place version is:

localmean!(dst, A, B) -> dst

See also localfilter!.

LocalFilters.offset — Function.

LocalFilters.offset(B) yields the index offset of the array of coefficients embedded in kernel B. That is, B[k] ≡ coefs(B)[k + offset(B)].

See also: LocalFilters.coefs.

LocalFilters.opening — Function.

closing(A, R)
opening(A, R)

respectively perform a closing or an opening of array A by the structuring element R. If not specified, R is a box of size 3 along all the dimensions of A. A closing is a dilation followed by an erosion, whereas an opening is an erosion followed by a dilation.

The in-place versions are:

closing!(dst, wrk, src, R)
opening!(dst, wrk, src, R)

which perform the operation on the source src and store the result in destination dst using wrk as a workspace array. These 3 arguments must be similar arrays, dst and src may be identical, but wrk must not be the same array as src or dst. The destination dst is returned.

See erode or dilate for the meaning of the arguments.

LocalFilters.opening! — Function.

closing(A, R)
opening(A, R)

respectively perform a closing or an opening of array A by the structuring element R. If not specified, R is a box of size 3 along all the dimensions of A. A closing is a dilation followed by an erosion, whereas an opening is an erosion followed by a dilation.

The in-place versions are:

closing!(dst, wrk, src, R)
opening!(dst, wrk, src, R)

which perform the operation on the source src and store the result in destination dst using wrk as a workspace array. These 3 arguments must be similar arrays, dst and src may be identical, but wrk must not be the same array as src or dst. The destination dst is returned.

See erode or dilate for the meaning of the arguments.

LocalFilters.strel — Function.

strel(T, A)

yields a structuring element suitable for mathematical morphology operations. The result is a Kernel whose elements have type T (which can be Bool or a floating-point type). Argument A can be a rectangular box or a Kernel with boolean elements.

If T is a floating-point type, then the result is a so-called flat structuring element whose coefficients are zero(T) inside the shape defined by A and -T(Inf) elsewhere.

See also: convertcoefs.

LocalFilters.strictfloor

LocalFilters.top_hat — Function.

top_hat(A, R)
top_hat(A, R, S)
bottom_hat(A, R)
bottom_hat(A, R, S)

Perform A summit/valley detection by applying a top-hat filter to array A. Argument R defines the structuring element for the feature detection. Optional argument S specifies the structuring element used to apply a smoothing to A prior to the top-hat filter. If R and S are specified as the radii of the structuring elements, then S should be smaller than R. For instance:

top_hat(bitmap, 3, 1)

may be used to detect text or lines in a bimap image.

The in-place versions:

top_hat!(dst, wrk, src, R)
bottom_hat!(dst, wrk, src, R)

apply the top-hat filter on the source src and store the result in the destination dst using wrk as a workspace array. These 3 arguments must be similar but different arrays. The destination dst is returned.

See also dilate, closing and morph_enhance.

LocalFilters.top_hat! — Function.

top_hat(A, R)
top_hat(A, R, S)
bottom_hat(A, R)
bottom_hat(A, R, S)

Perform A summit/valley detection by applying a top-hat filter to array A. Argument R defines the structuring element for the feature detection. Optional argument S specifies the structuring element used to apply a smoothing to A prior to the top-hat filter. If R and S are specified as the radii of the structuring elements, then S should be smaller than R. For instance:

top_hat(bitmap, 3, 1)

may be used to detect text or lines in a bimap image.

The in-place versions:

top_hat!(dst, wrk, src, R)
bottom_hat!(dst, wrk, src, R)

apply the top-hat filter on the source src and store the result in the destination dst using wrk as a workspace array. These 3 arguments must be similar but different arrays. The destination dst is returned.

See also dilate, closing and morph_enhance.