Docstrings

ZMQ.Context

ZMQ.DEALER

ZMQ.DOWNSTREAM

ZMQ.FORWARDER

ZMQ.IO_THREADS

ZMQ.IPV6

ZMQ.MAX_SOCKETS

ZMQ.MORE

ZMQ.Message

ZMQ.PAIR

ZMQ.POLLERR

ZMQ.POLLIN

ZMQ.POLLOUT

ZMQ.PUB

ZMQ.PULL

ZMQ.PUSH

ZMQ.QUEUE

ZMQ.REP

ZMQ.REQ

ZMQ.ROUTER

ZMQ.SNDMORE

ZMQ.STREAMER

ZMQ.SUB

ZMQ.Socket

ZMQ.StateError

ZMQ.UPSTREAM

ZMQ.XPUB

ZMQ.XREP

ZMQ.XREQ

ZMQ.XSUB

ZMQ.ZMQ

ZMQ.ZMQ_DONTWAIT

ZMQ.ZMQ_SNDMORE

ZMQ._Message

ZMQ._MessageOrRef

ZMQ._MessageRef

ZMQ.__init__

ZMQ.__zmq_getsockopt_rcvmore

ZMQ._context

ZMQ._ctx_new

ZMQ._get

ZMQ._get_affinity

ZMQ._get_backlog

ZMQ._get_events

ZMQ._get_fd

ZMQ._get_ipv4only

ZMQ._get_last_endpoint

ZMQ._get_linger

ZMQ._get_maxmsgsize

ZMQ._get_multicast_hops

ZMQ._get_rate

ZMQ._get_rcvbuf

ZMQ._get_rcvhwm

ZMQ._get_rcvmore

ZMQ._get_rcvtimeo

ZMQ._get_reconnect_ivl

ZMQ._get_reconnect_ivl_max

ZMQ._get_recovery_ivl

ZMQ._get_routing_id

ZMQ._get_sndbuf

ZMQ._get_sndhwm

ZMQ._get_sndtimeo

ZMQ._get_tcp_keepalive

ZMQ._get_tcp_keepalive_cnt

ZMQ._get_tcp_keepalive_idle

ZMQ._get_tcp_keepalive_intvl

ZMQ._get_type

ZMQ._recv!

ZMQ._send

ZMQ._set

ZMQ._set_affinity

ZMQ._set_backlog

ZMQ._set_ipv4only

ZMQ._set_linger

ZMQ._set_maxmsgsize

ZMQ._set_multicast_hops

ZMQ._set_rate

ZMQ._set_rcvbuf

ZMQ._set_rcvhwm

ZMQ._set_rcvtimeo

ZMQ._set_reconnect_ivl

ZMQ._set_reconnect_ivl_max

ZMQ._set_recovery_ivl

ZMQ._set_routing_id

ZMQ._set_sndbuf

ZMQ._set_sndhwm

ZMQ._set_sndtimeo

ZMQ._set_tcp_keepalive

ZMQ._set_tcp_keepalive_cnt

ZMQ._set_tcp_keepalive_idle

ZMQ._set_tcp_keepalive_intvl

ZMQ._set_type

ZMQ._zmq_getsockopt_rcvmore

ZMQ.bind

ZMQ.check_deps

ZMQ.connect

ZMQ.context — Function.

context()

Return the default ZMQ context (of type Context), initializing it if this has not been done already. (This context is automatically closed when Julia exits.)

ZMQ.ctxopts

ZMQ.depsjl_path

ZMQ.eval

ZMQ.gc_free_fn

ZMQ.gc_protect

ZMQ.gc_protect_cb

ZMQ.gc_protect_handle

ZMQ.get_affinity

ZMQ.get_backlog

ZMQ.get_events

ZMQ.get_fd

ZMQ.get_identity

ZMQ.get_ipv4only

ZMQ.get_linger

ZMQ.get_maxmsgsize

ZMQ.get_multicast_hops

ZMQ.get_rate

ZMQ.get_rcvbuf

ZMQ.get_rcvhwm

ZMQ.get_rcvmore

ZMQ.get_rcvtimeo

ZMQ.get_reconnect_ivl

ZMQ.get_reconnect_ivl_max

ZMQ.get_recovery_ivl

ZMQ.get_sndbuf

ZMQ.get_sndhwm

ZMQ.get_sndtimeo

ZMQ.get_tcp_keepalive

ZMQ.get_tcp_keepalive_cnt

ZMQ.get_tcp_keepalive_idle

ZMQ.get_tcp_keepalive_intvl

ZMQ.get_type

ZMQ.include

ZMQ.isfreed

ZMQ.ismore

ZMQ.jl_zmq_error_str

ZMQ.libzmq

ZMQ.msg_init

ZMQ.msg_recv

ZMQ.msg_send

ZMQ.propexpression

Sockets.recv — Function.

recv(socket::Socket) :: Message

Return a Message object representing a message received from a ZMQ Socket (without making a copy of the message data).

recv(socket::Socket, ::Type{T})

Receive a message of type T (typically a String, Vector{UInt8}, or isbits type) from a ZMQ Socket. (Makes a copy of the message data; you can alternatively use recv(socket) to work with zero-copy bytearray-like representation for large messages.)

Sockets.send — Function.

send(socket::Socket, data; more=false)

Send data over socket. A more=true keyword argument can be passed to indicate that data is a portion of a larger multipart message. data can be any isbits type, a Vector of isbits elements, a String, or a Message object to perform zero-copy sends of large arrays.

ZMQ.set

ZMQ.set_affinity

ZMQ.set_backlog

ZMQ.set_identity

ZMQ.set_ipv4only

ZMQ.set_linger

ZMQ.set_maxmsgsize

ZMQ.set_multicast_hops

ZMQ.set_rate

ZMQ.set_rcvbuf

ZMQ.set_rcvhwm

ZMQ.set_rcvtimeo

ZMQ.set_reconnect_ivl

ZMQ.set_reconnect_ivl_max

ZMQ.set_recovery_ivl

ZMQ.set_sndbuf

ZMQ.set_sndhwm

ZMQ.set_sndtimeo

ZMQ.set_tcp_keepalive

ZMQ.set_tcp_keepalive_cnt

ZMQ.set_tcp_keepalive_idle

ZMQ.set_tcp_keepalive_intvl

ZMQ.set_type

ZMQ.show

ZMQ.sockprops

ZMQ.subscribe

ZMQ.subscribe_

ZMQ.term

ZMQ.unsafe_copy

ZMQ.unsubscribe

ZMQ.unsubscribe_

ZMQ.version

ZMQ.zmq_errno