Docstrings

SymmetricTensors.*

SymmetricTensors.+

SymmetricTensors.-

SymmetricTensors./

SymmetricTensors.ArrayNArrays

SymmetricTensors.SymmetricTensor — Type.

Type constructor

frame - stores ArrayNArrays{T,N} bls - Int, size of ordinary block bln - Int, number of blocks datasize - Int, size of data stored (in each direction the same) sqr - Bool, is the last block size a same as ordinary's block size

SymmetricTensors.SymmetricTensors

SymmetricTensors.arraynarrays

SymmetricTensors.convert

SymmetricTensors.diag — Function.

diag(st::SymmetricTensor{N})

Return vector of floats, the super-diag of st

SymmetricTensors.eval

SymmetricTensors.fixpointperms — Function.

fixpointperms(j::NTuple{N, Int}) where N

Returns Vactor{Vector}, a fix point permutation of given multiindex

TODO: this is a naive implementation

SymmetricTensors.frtest — Function.

frtest(data::ArrayNArrays{T,N})

Returns assertion error if: all sizes of nullable array not equal, at least some undergiagonal block not null, some blocks (not last) not squared, some diagonal blocks not symmetric.

SymmetricTensors.getblock — Function.

getblock(st::SymmetricTensor, i::Tuple)

Returns a block from Symmetric Tensor, works for all multi-indices also not sorted

SymmetricTensors.getblockunsafe — Function.

getblockunsafe(st::SymmetricTensor, i::Tuple)

Returns a block from Symmetric Tensor, unsafe works only if multi-index is sorted

SymmetricTensors.include

SymmetricTensors.ind2range — Function.

ind2range(i::Int, bls::Int, dats::Int)

Returns a range given index i, size of a block and size of data

julia> ind2range(2,3,5)
4:5

SymmetricTensors.issymetric — Function.

issymetric(ar::Array{N}, atol::Float64)

Returns: Assertion Error if not symmetric given a tolerance.

julia> A = reshape(collect(1.:8.), 2, 2, 2);

julia> julia> issymetric(A)
ERROR: AssertionError: not symmetric

SymmetricTensors.pyramidindices — Function.

pyramidindices(dims::Int, tensize::Int)

julia> pyramidindices(2,3)
6-element Array{Tuple{Int64,Int64},1}:
 (1,1)
 (1,2)
 (1,3)
 (2,2)
 (2,3)
 (3,3)

Base.rand — Function.

rand(SymmetricTensor{T, N}, n::Int, b::Int = 2)

Returns N-dimensional random SymmetricTensor with elements of type T drawn from uniform distribution on [0,1), n denotes data size and b denotes block size.

SymmetricTensors.randblock — Function.

randblock(::Type{T}, dims::NTuple{N, Int}, j::NTuple{N, Int})

Returns a block of size dims and position j by a uniformly distributed random number of type T

SymmetricTensors.randsymarray — Function.

randsymarray(T, dim ::Int, N::Int = 4)

Returns N-dimmensional random super-symmetric array with elements of type T drawn from uniform distribution on [0,1), dim denotes data size.

randsymarray(T, dim::Int, N::Int = 4)

Returns N-dimmensional random super-symmetric array with Float64 elements drawn from uniform distribution on [0,1), dim denotes data size.

Base.setindex! — Function.

setindex!(st::SymmetricTensor{T,N}, x::T,  mulind::Int...)

Change a SymmetricTensors value at the given multi-index

SymmetricTensors.setindexunsafe! — Function.

setindexunsafe!(st::SymmetricTensor{T,N}, x::T,  mulind::Int...)

Unsafe change a SymmetricTensors value at the given multi-index

SymmetricTensors.sizetest — Function.

sizetest(dats::Int, bls::Int)

Returns: DimensionMismatch if blocks size is grater than data size.

julia> SymmetricTensors.sizetest(2,3)
ERROR: DimensionMismatch("bad block size 3 > 2")

SymmetricTensors.unfold — Function.

unfold(ar::Array{N}, mode::Int)

Returns an matrix being an unfold of N dims array in given mode.

julia> A = reshape(collect(1.:8.), 2, 2, 2);

julia> unfold(A, 1)
2×4 Array{Float64,2}:
 1.0  3.0  5.0  7.0
 2.0  4.0  6.0  8.0

 julia> unfold(A, 2)
 2×4 Array{Float64,2}:
  1.0  2.0  5.0  6.0
  3.0  4.0  7.0  8.0

  julia> unfold(A, 3)
  2×4 Array{Float64,2}:
   1.0  2.0  3.0  4.0
   5.0  6.0  7.0  8.0