Docstrings

PlyIO.ArrayProperty — Type.

ArrayProperty(name, T)

A ply property $T $name, modelled as an abstract vector, with a name which can be retrieved using plyname().

PlyIO.Format

PlyIO.Format_ascii

PlyIO.Format_binary_big

PlyIO.Format_binary_little

PlyIO.ListProperty — Type.

ListProperty(name, S, T)
ListProperty(name, list_of_vectors)

A ply property list $S $T $name, modelled as a abstract vector of vectors, with a name which can be retrieved using plyname().

PlyIO.Ply — Type.

Ply()

Container for the contents of a ply file. This type directly models the contents of the header. Ply elements and comments can be added using push!(), elements can be iterated over with the standard iterator interface, and looked up by indexing with a string.

PlyIO.PlyComment — Type.

PlyComment(string)

A ply comment.

PlyIO.PlyElement — Type.

PlyElement(name, [len | props...])

Construct a ply element $name $len, containing a list of properties with a name which can be retrieved using plyname. Properties can be accessed with the array interface, or looked up by indexing with a string.

The expected length len is used if it is set, otherwise the length shared by the property vectors is used.

PlyIO.PlyIO

PlyIO.PlyNativeType

PlyIO.PropNameList

PlyIO._host_is_little_endian

PlyIO.eval

PlyIO.include

PlyIO.load_ply — Function.

load_ply(file)

Load data from a ply file and return a Ply datastructure. file may either be a file name or an open stream.

PlyIO.parse_ascii

PlyIO.ply_type

PlyIO.ply_type_name

PlyIO.plyname — Function.

plyname(data)

Return the name that data is associated with when serialized in a ply file

PlyIO.read_ascii_value!

PlyIO.read_binary_value!

PlyIO.read_binary_values!

PlyIO.read_header

PlyIO.save_ply — Function.

save_ply(ply::Ply, file; [ascii=false])

Save data from Ply data structure into file which may be a filename or an open stream. The file will be native endian binary, unless the keyword argument ascii is set to true.

PlyIO.write_ascii_value

PlyIO.write_binary_value

PlyIO.write_binary_values

PlyIO.write_header

PlyIO.write_header_field