Docstrings

ArgCheck.@argcheck — Macro.

@argcheck

Check invariants on function arguments and produce a nice exception message if they are violated. Usage is as follows:

function myfunction(k,n,A,B)
    @argcheck k > n
    @argcheck size(A) == size(B) DimensionMismatch
    @argcheck det(A) < 0 DomainError()
    # doit
end

See also @check.

ArgCheck.@check — Macro.

@check

Check that a condition holds and produce a nice exception message, if it does not. Usage is as follows:

@check k > n
@check size(A) == size(B) DimensionMismatch
@check det(A) < 0 DomainError()

See also @argcheck.

ArgCheck.AbstractCheckFlavor

ArgCheck.AbstractCodeFlavor

ArgCheck.AbstractErrorInfo

ArgCheck.ArgCheck

ArgCheck.ArgCheckFlavor

ArgCheck.CallErrorInfo

ArgCheck.CallFlavor

ArgCheck.CheckError

ArgCheck.CheckFlavor

ArgCheck.Checker

ArgCheck.ComparisonErrorInfo

ArgCheck.ComparisonFlavor

ArgCheck.FallbackErrorInfo

ArgCheck.FallbackFlavor

ArgCheck.build_error

ArgCheck.check

ArgCheck.default_exception_type

ArgCheck.error_message

ArgCheck.eval

ArgCheck.expr_error_block

ArgCheck.fancy_error_message

ArgCheck.include

ArgCheck.is_simple_call

ArgCheck.pretty_string