Docstrings

NumericalMethodsforEngineers./

NumericalMethodsforEngineers.@compat

NumericalMethodsforEngineers.@dotcompat

NumericalMethodsforEngineers.@functorize

NumericalMethodsforEngineers.@sym

NumericalMethodsforEngineers.ARPACKException

NumericalMethodsforEngineers.AbstractSparseArray

NumericalMethodsforEngineers.AbstractSparseMatrix

NumericalMethodsforEngineers.AbstractSparseVector

NumericalMethodsforEngineers.Adjoint

NumericalMethodsforEngineers.BLAS

NumericalMethodsforEngineers.Bidiagonal

NumericalMethodsforEngineers.BunchKaufman

NumericalMethodsforEngineers.Cholesky

NumericalMethodsforEngineers.CholeskyPivoted

NumericalMethodsforEngineers.Compat

NumericalMethodsforEngineers.Diagonal

NumericalMethodsforEngineers.Eigen

NumericalMethodsforEngineers.Factorization

NumericalMethodsforEngineers.GeneralizedEigen

NumericalMethodsforEngineers.GeneralizedSVD

NumericalMethodsforEngineers.GeneralizedSchur

NumericalMethodsforEngineers.Hermitian

NumericalMethodsforEngineers.Hessenberg

NumericalMethodsforEngineers.I

NumericalMethodsforEngineers.LAPACK

NumericalMethodsforEngineers.LAPACKException

NumericalMethodsforEngineers.LDLt

NumericalMethodsforEngineers.LQ

NumericalMethodsforEngineers.LU

NumericalMethodsforEngineers.LinearAlgebra

NumericalMethodsforEngineers.LowerTriangular

NumericalMethodsforEngineers.NumericalMethodsforEngineers

NumericalMethodsforEngineers.PosDefException

NumericalMethodsforEngineers.QR

NumericalMethodsforEngineers.QRPivoted

NumericalMethodsforEngineers.RankDeficientException

NumericalMethodsforEngineers.SVD

NumericalMethodsforEngineers.Schur

NumericalMethodsforEngineers.SingularException

NumericalMethodsforEngineers.SkylineMatrix — Type.

SkylineMatrix type for symmetric matrices

Type to hold skyline matrices with element types Tv

sv: Skyline vector

kdiag: Diagonal elements in sv.

sm: SkylineMatrix.

Constructor

sm = SkylineMatrix{Tv}(kdiag::Vector{Int}, sv::Vector{Tv})

Arguments

* `kdiag`           : Vector specifying location of diagonal element in sv. 
* `sv`              : Non-zero values in upper triangular matrix.

Additional methods

* `size(sm)`        : Return size of full matrix => (n, n).
* `==(sm1, sm2)`    : Compare 2 SkylineMatrices for equality.
* `copy(sm)`        : Create a copy.
* `sparse(sm)`      : Convert to a SparseMatrixCSC.
* `full(sm)`        : Convert to a full matrix.

* `convert(::Type{NumericalMethodsforEngineers.SkylineMatrix}, kdiag::Vector{Int}, sv::Vector{Tv})`
                    : Default constructor

* `fromskyline(kdiag::Vector{Int}, sv::Vector{Tv})`
                    : Convert kdiag,sv to full matrix.

NumericalMethodsforEngineers.SparseArrays

NumericalMethodsforEngineers.SparseMatrixCSC

NumericalMethodsforEngineers.SparseVector

NumericalMethodsforEngineers.SymTridiagonal

NumericalMethodsforEngineers.Symmetric

NumericalMethodsforEngineers.SymmetricBandedMatrix

NumericalMethodsforEngineers.Transpose

NumericalMethodsforEngineers.Tridiagonal

NumericalMethodsforEngineers.UniformScaling

NumericalMethodsforEngineers.UnitLowerTriangular

NumericalMethodsforEngineers.UnitUpperTriangular

NumericalMethodsforEngineers.UpperTriangular

NumericalMethodsforEngineers.\

NumericalMethodsforEngineers.adjoint

NumericalMethodsforEngineers.adjoint!

NumericalMethodsforEngineers.axpby!

NumericalMethodsforEngineers.axpy!

NumericalMethodsforEngineers.bicgstab

NumericalMethodsforEngineers.bicgstab!

NumericalMethodsforEngineers.bigcstab

NumericalMethodsforEngineers.blockdiag

NumericalMethodsforEngineers.bunchkaufman

NumericalMethodsforEngineers.bunchkaufman!

NumericalMethodsforEngineers.checkit

NumericalMethodsforEngineers.chobac!

NumericalMethodsforEngineers.cholesky

NumericalMethodsforEngineers.cholesky!

NumericalMethodsforEngineers.cholin!

NumericalMethodsforEngineers.cond

NumericalMethodsforEngineers.condskeel

NumericalMethodsforEngineers.copy_transpose!

NumericalMethodsforEngineers.copyto!

NumericalMethodsforEngineers.cross

NumericalMethodsforEngineers.det

NumericalMethodsforEngineers.diag

NumericalMethodsforEngineers.diagind

NumericalMethodsforEngineers.diagm

NumericalMethodsforEngineers.dot

NumericalMethodsforEngineers.droptol!

NumericalMethodsforEngineers.dropzeros

NumericalMethodsforEngineers.dropzeros!

NumericalMethodsforEngineers.eigen

NumericalMethodsforEngineers.eigen!

NumericalMethodsforEngineers.eigmax

NumericalMethodsforEngineers.eigmin

NumericalMethodsforEngineers.eigvals

NumericalMethodsforEngineers.eigvals!

NumericalMethodsforEngineers.eigvecs

NumericalMethodsforEngineers.euler

NumericalMethodsforEngineers.eval

NumericalMethodsforEngineers.factorize

NumericalMethodsforEngineers.findnz

NumericalMethodsforEngineers.fromskyline — Function.

Convert the pair kdiag, sv to a full matrix

Function

m = fromskyline{Tv}(kdiag::Vector{Int}, sv::Vector{Tv})

Arguments

* `kdiag`           : Vector specifying location of diagonal element in sv. 
* `sv`              : Non-zero values in upper triangular matrix.

NumericalMethodsforEngineers.full

NumericalMethodsforEngineers.getsymata

NumericalMethodsforEngineers.givens

NumericalMethodsforEngineers.hessenberg

NumericalMethodsforEngineers.hessenberg!

NumericalMethodsforEngineers.include

NumericalMethodsforEngineers.isdiag

NumericalMethodsforEngineers.ishermitian

NumericalMethodsforEngineers.isposdef

NumericalMethodsforEngineers.isposdef!

NumericalMethodsforEngineers.issparse

NumericalMethodsforEngineers.issuccess

NumericalMethodsforEngineers.issymmetric

NumericalMethodsforEngineers.istril

NumericalMethodsforEngineers.istriu

NumericalMethodsforEngineers.kron

NumericalMethodsforEngineers.lagrangianpolynomial

NumericalMethodsforEngineers.ldiv!

NumericalMethodsforEngineers.ldlfor!

NumericalMethodsforEngineers.ldlt

NumericalMethodsforEngineers.ldlt!

NumericalMethodsforEngineers.ldltfac!

NumericalMethodsforEngineers.lmul!

NumericalMethodsforEngineers.logabsdet

NumericalMethodsforEngineers.logdet

NumericalMethodsforEngineers.lowrankdowndate

NumericalMethodsforEngineers.lowrankdowndate!

NumericalMethodsforEngineers.lowrankupdate

NumericalMethodsforEngineers.lowrankupdate!

NumericalMethodsforEngineers.lq

NumericalMethodsforEngineers.lq!

NumericalMethodsforEngineers.lu

NumericalMethodsforEngineers.lu!

NumericalMethodsforEngineers.lufac

NumericalMethodsforEngineers.lyap

NumericalMethodsforEngineers.mid_point_euler

NumericalMethodsforEngineers.modified_euler

NumericalMethodsforEngineers.mul!

NumericalMethodsforEngineers.nmex

NumericalMethodsforEngineers.nnz

NumericalMethodsforEngineers.nonzeros

NumericalMethodsforEngineers.norm

NumericalMethodsforEngineers.normalize

NumericalMethodsforEngineers.normalize!

NumericalMethodsforEngineers.nullspace

NumericalMethodsforEngineers.nzrange

NumericalMethodsforEngineers.opnorm

NumericalMethodsforEngineers.ordschur

NumericalMethodsforEngineers.ordschur!

NumericalMethodsforEngineers.permute

NumericalMethodsforEngineers.pinv

NumericalMethodsforEngineers.qr

NumericalMethodsforEngineers.qr!

NumericalMethodsforEngineers.rank

NumericalMethodsforEngineers.rdiv!

NumericalMethodsforEngineers.rmul!

NumericalMethodsforEngineers.rowvals

NumericalMethodsforEngineers.runga_kutta_4

NumericalMethodsforEngineers.schur

NumericalMethodsforEngineers.schur!

NumericalMethodsforEngineers.setsymata

NumericalMethodsforEngineers.shootingmethod — Function.

Function shootingmethod() for 2nd order BVP ODEs.

Arguments to shootingmethod(f, bvs, initg, steps, tol, limit):

f::Function : Function to be estimated (see below).

bvs::Vector{Float64} : Vector of boundary values [xa, ya, xb, yb].

initg::Matrix{Float64}: Vector of nitial gradients [a0, a1].

steps::Int64 : Number of calculations steps.

tol::Float64 : Tolerance for convergence.

limit::Int64 : Max number of iterations.

Example of how to define the function f(x, y) for y'' = 3x^3+4y:

Decompose in two first order equations:

dy[1] = y[2]

dy[2] = 3x^3 + 4y[1]

function f(x::Float64, y::Vector{Float64})
  [y[2], 3x^3 +4y[1]]
end

NumericalMethodsforEngineers.spabac!

NumericalMethodsforEngineers.sparin!

NumericalMethodsforEngineers.sparse

NumericalMethodsforEngineers.sparsevec

NumericalMethodsforEngineers.spdiagm

NumericalMethodsforEngineers.sprand

NumericalMethodsforEngineers.sprandn

NumericalMethodsforEngineers.spzeros

NumericalMethodsforEngineers.subbac!

NumericalMethodsforEngineers.subfor!

NumericalMethodsforEngineers.svd

NumericalMethodsforEngineers.svd!

NumericalMethodsforEngineers.svdvals

NumericalMethodsforEngineers.svdvals!

NumericalMethodsforEngineers.sylvester

NumericalMethodsforEngineers.symeval

NumericalMethodsforEngineers.tosymmetricbandedmatrix — Function.

Convert symmetric AbstractMatrix to a SymmetricBandedMatrix

Function

bm = tosymmetricbandedmatrix{Tv}(hbw::Int, am::AbstractMatrix{Tv})
or
bm = tosymmetricbandedmatrix{Tv}(am::AbstractMatrix{Tv})

Arguments

* `hbw`             : Half bandwidth. 
* `am`              : Symmetric AbstractMatrix
                      (if no hbw is specified, hbw will be derived from the matrix)

NumericalMethodsforEngineers.tr

NumericalMethodsforEngineers.transpose

NumericalMethodsforEngineers.transpose!

NumericalMethodsforEngineers.tril

NumericalMethodsforEngineers.tril!

NumericalMethodsforEngineers.triu

NumericalMethodsforEngineers.triu!

NumericalMethodsforEngineers.×

NumericalMethodsforEngineers.⋅