Docstrings

ChangePrecision.*

ChangePrecision.+

ChangePrecision.-

ChangePrecision./

ChangePrecision.@changeprecision — Macro.

@changeprecision T expression

Change the "default" precision in the given expression to the floating-point type T.

This changes floating-point literals, integer expressions like 1/3, random-number functions like rand, and matrix constructors like ones to default to the new type T.

For example,

@changeprecision Float32 begin
    x = 7.3
    y = 1/3
    z = rand() .+ ones(3,4)
end

uses Float32 precision for all of the expressions in begin ... end.

ChangePrecision.ChangePrecision — Module.

The ChangePrecision module exports a macro @changeprecision T expression that changes the "default" floating-point precision in a given expression to a new floating-point type T.

ChangePrecision.HWInt

ChangePrecision.IntLike

ChangePrecision.IntRatLike

ChangePrecision.Promotable

ChangePrecision.PromotableNoRat

ChangePrecision.RatLike

ChangePrecision.\

ChangePrecision.^

ChangePrecision.abs

ChangePrecision.acos

ChangePrecision.acosd

ChangePrecision.acosh

ChangePrecision.acot

ChangePrecision.acotd

ChangePrecision.acoth

ChangePrecision.acsc

ChangePrecision.acscd

ChangePrecision.acsch

ChangePrecision.angle

ChangePrecision.asec

ChangePrecision.asecd

ChangePrecision.asech

ChangePrecision.asin

ChangePrecision.asind

ChangePrecision.asinh

ChangePrecision.atan

ChangePrecision.atan2

ChangePrecision.atand

ChangePrecision.atanh

ChangePrecision.binaryfuncs

ChangePrecision.bunchkaufman

ChangePrecision.cbrt

ChangePrecision.changefuncs

ChangePrecision.changeprecision

ChangePrecision.cholesky

ChangePrecision.cis

ChangePrecision.complexfuncs

ChangePrecision.cond

ChangePrecision.condskeel

ChangePrecision.cor

ChangePrecision.cos

ChangePrecision.cosc

ChangePrecision.cosd

ChangePrecision.cosh

ChangePrecision.cospi

ChangePrecision.cot

ChangePrecision.cotd

ChangePrecision.coth

ChangePrecision.cov

ChangePrecision.csc

ChangePrecision.cscd

ChangePrecision.csch

ChangePrecision.deg2rad

ChangePrecision.det

ChangePrecision.eigen

ChangePrecision.eigfact

ChangePrecision.eigmax

ChangePrecision.eigmin

ChangePrecision.eigvals

ChangePrecision.eigvecs

ChangePrecision.esym

ChangePrecision.eval

ChangePrecision.exp

ChangePrecision.exp10

ChangePrecision.exp2

ChangePrecision.expm1

ChangePrecision.factorize

ChangePrecision.float

ChangePrecision.hessenberg

ChangePrecision.hypot

ChangePrecision.include

ChangePrecision.intfuncs

ChangePrecision.inv

ChangePrecision.ldlt

ChangePrecision.linalgfuncs

ChangePrecision.literal_pow

ChangePrecision.log

ChangePrecision.log10

ChangePrecision.log1p

ChangePrecision.log2

ChangePrecision.logabsdet

ChangePrecision.logdet

ChangePrecision.lq

ChangePrecision.lu

ChangePrecision.lyap

ChangePrecision.matfuncs

ChangePrecision.mean

ChangePrecision.median

ChangePrecision.norm

ChangePrecision.normalize

ChangePrecision.nullspace

ChangePrecision.ones

ChangePrecision.opnorm

ChangePrecision.pinv

ChangePrecision.qr

ChangePrecision.rad2deg

ChangePrecision.rand

ChangePrecision.randexp

ChangePrecision.randfuncs

ChangePrecision.randn

ChangePrecision.schur

ChangePrecision.sec

ChangePrecision.secd

ChangePrecision.sech

ChangePrecision.sin

ChangePrecision.sinc

ChangePrecision.sind

ChangePrecision.sinh

ChangePrecision.sinpi

ChangePrecision.sqrt

ChangePrecision.statfuncs

ChangePrecision.std

ChangePrecision.stdm

ChangePrecision.svd

ChangePrecision.svdvals

ChangePrecision.sylvester

ChangePrecision.tan

ChangePrecision.tand

ChangePrecision.tanh

ChangePrecision.tofloat

ChangePrecision.var

ChangePrecision.varm

ChangePrecision.zeros