Docstrings

SyntaxTree.@genfun — Macro.

@genfun(expr, args)

Returns an anonymous function based on the given expr and args.

julia> @genfun x^2+y^2 x y

SyntaxTree.SyntaxTree

SyntaxTree.abs — Function.

SyntaxTree.abs(expr)

Apply abs to the expression recursively.

SyntaxTree.alg — Function.

alg(expr,f=:(1+ϵ))

Recursively insert a machine epsilon bound (1+ϵ) per call in expr.

SyntaxTree.callcount — Function.

callcount(expr)

Returns a count of the call operations in expr.

SyntaxTree.eval

SyntaxTree.expravg — Function.

expravg(expr)

Recursively obtains the number of scalars in an expression, the average of those scalars logarithms, the number of non-trivial exponents in the expression, and the average of the exponents.

SyntaxTree.exprdev — Function.

exprdev(expr)

Returns the standard deviation of the logarithm of the scalars in an expression.

SyntaxTree.exprval — Function.

exprval(expr)

Returns the expression value and other characteristics as defined in "Optimal polynomial characteristic methods" by Michael Reed in 2018. This value can be used to order polynomial forms, with lower values being more optimal and efficient for computation.

The first value is the expression value, the second is the number of operations, the third is the standard deviation of the logarithms of the scalars, the fourth is the average of the logarithms of the scalars, and the fifth is the average of the exponents.

SyntaxTree.genfun — Function.

genfun(expr, args)

Returns an anonymous function based on the given expr and args.

julia> genfun(:(x^2+y^2),[:x,:y])
julia> genfun(:(x^2+y^2),(:x,:y))
julia> genfun(:(x^2+y^2),:x,:y)

SyntaxTree.include

SyntaxTree.linefilter

SyntaxTree.linefilter! — Function.

linefilter!(::Expr)

Recursively filters out :LineNumberNode from Expr objects.

SyntaxTree.sub — Function.

sub(T::DataType,expr::Expr)

Make a substitution to convert numerical values inside an Expr to type T.