Docstrings

MirroredArrayViews.MirroredArrayView — Type.

MirroredArrayView(arr::AbstractArray, dims...)

Construct a view of array arr mirrored along each dimension in dims. Repeated dimensions will be ignored. The mirrored dimensions become part of the type of the view and no data is copied.

MirroredArrayViews.MirroredArrayViews

MirroredArrayViews.eval

MirroredArrayViews.include

MirroredArrayViews.mirror_indices