Docstrings

BoltzmannMachinesPlots.BMs

BoltzmannMachinesPlots.BoltzmannMachinesPlots — Module.

Contains all plotting functions for displaying information collected in module BoltzmannMachines. Most important function is plotevaluation.

BoltzmannMachinesPlots.checkdata

BoltzmannMachinesPlots.crossvalidationcurve

BoltzmannMachinesPlots.eval

BoltzmannMachinesPlots.extractaisdata

BoltzmannMachinesPlots.extractevaluationdata

BoltzmannMachinesPlots.getvalue

BoltzmannMachinesPlots.include

BoltzmannMachinesPlots.plotcurvebundles

BoltzmannMachinesPlots.plotestimatedprob — Function.

Plots the information about the estimated lower bound of the log probability that has been gathered while training a BMs.

BoltzmannMachinesPlots.plotevaluation — Function.

plotevaluation(monitor; ...)
plotevaluation(monitor, evaluationkey; ...)

Plots a curve that shows the values of the evaluation contained in the monitor and specified by the evaluationkey over the course of the training epochs. If no evaluationkey is specified, the evaluation type of the first monitor element is used.

Optional keyword argument sdrange: For evaluations with keys BoltzmannMachines.monitorloglikelihood and BoltzmannMachines.monitorlogproblowerbound, there is additional information about the standard deviation of the estimator. With the parameter sdrange, it is possible to display this information as a ribbon around the curve. The ribbon indicates the area around the curve that contains the values that deviate at maximum sdrange times the standard deviation from the estimator. Default value for sdrange is 2.0.

BoltzmannMachinesPlots.plottitledict

BoltzmannMachinesPlots.scatter

BoltzmannMachinesPlots.scatterhidden — Function.

scatterhidden(bm, x; ...)
scatterhidden(h; ...)

Creates a scatter plot of the logarithmized activation potential of hidden nodes, similar to a PCA plot. The activation is either induced by the dataset x in the Boltzmann machine bm or it is directly specified as matrix h.

Optional keyword arguments:

hiddennodes: Tuple of integers, default (1,2), selecting the first two nodes of the (last) hidden layer.

labels: a vector containing string labels for each of the data points