Docstrings

Plots.@animate — Macro.

Collect one frame per for-block iteration and return an Animation object.

Example:

p = plot(1)
anim = @animate for x=0:0.1:5
    push!(p, 1, sin(x))
end
gif(anim)

Plots.@colorant_str

Plots.@deps

Plots.@gif — Macro.

Builds an Animation using one frame per loop iteration, then create an animated GIF.

Example:

  p = plot(1)
  @gif for x=0:0.1:5
    push!(p, 1, sin(x))
  end

Plots.@init_backend

Plots.@layout

Plots.@recipe

Plots.@series

Plots.@shorthands

Plots.@userplot

Plots.ABGR

Plots.ADIN99

Plots.ADIN99d

Plots.ADIN99o

Plots.AGray

Plots.AGray32

Plots.AHSI

Plots.AHSL

Plots.AHSV

Plots.ALCHab

Plots.ALCHuv

Plots.ALMS

Plots.ALab

Plots.ALuv

Plots.AMat

Plots.ARGB

Plots.ARGB32

Plots.AVec

Plots.AXYZ

Plots.AYCbCr

Plots.AYIQ

Plots.AbstractBackend

Plots.AbstractGray

Plots.AbstractLayout

Plots.AbstractPlot

Plots.AbstractRGB

Plots.AbstractSurface

Plots.AlphaColor

Plots.AnimatedGif — Type.

Wraps the location of an animated gif so that it can be displayed

Plots.Animation — Type.

Represents an animation object

Plots.Arrow

Plots.Axis

Plots.AxyY

Plots.BGR

Plots.BGRA

Plots.BezierCurve — Type.

create a BezierCurve for plotting

Plots.Brush

Plots.CIE1931JV_CMF

Plots.CIE1931J_CMF

Plots.CIE1931_CMF

Plots.CIE1964_CMF

Plots.CURRENT_BACKEND

Plots.CURRENT_PLOT

Plots.Color

Plots.Color3

Plots.ColorAlpha

Plots.ColorGradient

Plots.ColorTypes

Plots.Colorant

Plots.ColorantNormed

Plots.Colors

Plots.CurrentBackend

Plots.CurrentPlot

Plots.DD

Plots.DE_2000

Plots.DE_94

Plots.DE_AB

Plots.DE_BFD

Plots.DE_CMC

Plots.DE_DIN99

Plots.DE_DIN99d

Plots.DE_DIN99o

Plots.DE_JPC79

Plots.DIN99

Plots.DIN99A

Plots.DIN99d

Plots.DIN99dA

Plots.DIN99o

Plots.DIN99oA

Plots.DPI

Plots.DebugMode

Plots.EachAnn

Plots.EmptyLayout

Plots.Extrema

Plots.Font

Plots.Formatted — Type.

Represents data values with formatting that should apply to the tick labels.

Plots.Fractional

Plots.FrameIterator

Plots.FuncOrFuncs

Plots.GR

Plots.GRBackend

Plots.Gray

Plots.Gray24

Plots.GrayA

Plots.GridLayout

Plots.GroupBy — Type.

A special type that will break up incoming data into groups, and allow for easier creation of grouped plots

Plots.HDF5Backend

Plots.HSB

Plots.HSI

Plots.HSIA

Plots.HSL

Plots.HSLA

Plots.HSV

Plots.HSVA

Plots.InputWrapper

Plots.InspectDRBackend

Plots.KW

Plots.LCHab

Plots.LCHabA

Plots.LCHuv

Plots.LCHuvA

Plots.LMS

Plots.LMSA

Plots.Lab

Plots.LabA

Plots.Luv

Plots.LuvA

Plots.MM_PER_INCH

Plots.MM_PER_PX

Plots.MSC

Plots.MixedMeasures

Plots.NoBackend

Plots.OHLC — Type.

Represent Open High Low Close data (used in finance)

Plots.P2

Plots.P3

Plots.PGFPlotsBackend

Plots.PX_PER_INCH

Plots.Plot

Plots.PlotExample — Type.

Holds all data needed for a documentation example... header, description, and plotting expression (Expr)

Plots.PlotMeasures

Plots.PlotText — Type.

Wrap a string with font info

Plots.PlotThemes

Plots.PlotUtils

Plots.PlotlyBackend

Plots.PlotlyJSBackend

Plots.Plots

Plots.PlotsDisplay

Plots.PortfolioComposition

Plots.PyPlotBackend

Plots.RGB

Plots.RGB1

Plots.RGB24

Plots.RGB4

Plots.RGBA

Plots.RecipeData

Plots.RecipesBase

Plots.RootLayout

Plots.Segments — Type.

Build line segments for plotting

Plots.SegmentsIterator

Plots.Series

Plots.SeriesAnnotations

Plots.Shape — Type.

Shape(x, y)
Shape(vertices)

Construct a polygon to be plotted

Plots.ShowGradient

Plots.ShowLibrary

Plots.ShowTheme

Plots.SliceIt

Plots.Spy

Plots.Stroke

Plots.Subplot

Plots.SubplotMap

Plots.Surface — Type.

represents a contour or surface mesh

Plots.SurfaceFunction — Type.

For the case of representing a surface as a function of x/y... can possibly avoid allocations.

Plots.TicksArgs

Plots.Transparent3

Plots.TransparentColor

Plots.TransparentGray

Plots.TransparentRGB

Plots.UnicodePlotsBackend

Plots.Volume

Plots.XYZ

Plots.XYZA

Plots.YCbCr

Plots.YCbCrA

Plots.YIQ

Plots.YIQA

Plots.ZValues — Type.

type which represents z-values for colors and sizes (and anything else that might come up)

Plots._3dTypes

Plots.__init__

Plots._add_defaults!

Plots._add_errorbar_kw

Plots._add_markershape

Plots._add_smooth_kw

Plots._add_the_series

Plots._allAxes

Plots._allFramestyles

Plots._allGridArgs

Plots._allGridSyms

Plots._allMarkers

Plots._allScales

Plots._allShowaxisArgs

Plots._allShowaxisSyms

Plots._allStyles

Plots._allTypes

Plots._all_args

Plots._all_defaults

Plots._already_warned

Plots._animate

Plots._animation_examples

Plots._apply_type_recipe

Plots._arg_desc

Plots._attribute_defaults

Plots._auto_binning_nbins

Plots._axesAliases

Plots._axis_defaults

Plots._axis_defaults_byletter

Plots._backendPackage

Plots._backendSymbol

Plots._backendType

Plots._backend_instance

Plots._backend_skips

Plots._backends

Plots._bar_width

Plots._base_supported_args

Plots._before_layout_calcs

Plots._best_html_output_type

Plots._bin_centers

Plots._binbarlike_baseline

Plots._cbar_width

Plots._color_functions

Plots._create_backend_figure

Plots._current_plots_version

Plots._cycle

Plots._debugMode

Plots._default_backends

Plots._deprecated_backends

Plots._display

Plots._do_plot_show

Plots._document_argument

Plots._examples

Plots._expand_limits

Plots._expand_seriestype_array

Plots._expand_subplot_extrema

Plots._filter_input_data!

Plots._filternans

Plots._fmt_paragraph

Plots._framestyleAliases

Plots._get_defaults

Plots._get_showtheme_args

Plots._gr_attr

Plots._gr_gradient_alpha

Plots._gr_marker

Plots._gr_mimeformats

Plots._gr_point_mult

Plots._gr_scale

Plots._gr_seriestype

Plots._gr_style

Plots._gr_thickness_scaling

Plots._gr_wstype

Plots._gr_wstype_default

Plots._hdf5_attr

Plots._hdf5_marker

Plots._hdf5_scale

Plots._hdf5_seriestype

Plots._hdf5_style

Plots._heatmap_edges

Plots._hist_edge

Plots._hist_edges

Plots._hist_norm_mode

Plots._histogram_like

Plots._initial_axis_defaults

Plots._initial_defaults

Plots._initial_fontsizes

Plots._initialize_backend

Plots._initialize_subplot

Plots._initialized_backends

Plots._inspectdr_attr

Plots._inspectdr_marker

Plots._inspectdr_scale

Plots._inspectdr_seriestype

Plots._inspectdr_style

Plots._inv_scale_funcs

Plots._is_positive

Plots._keyAliases

Plots._label_func

Plots._line_like

Plots._logScaleBases

Plots._logScales

Plots._make_hist

Plots._markerAliases

Plots._match_map

Plots._match_map2

Plots._override_seriestype_check

Plots._pgfplots_attr

Plots._pgfplots_marker

Plots._pgfplots_scale

Plots._pgfplots_seriestype

Plots._pgfplots_style

Plots._plot!

Plots._plot_defaults

Plots._plot_setup

Plots._plotly_attr

Plots._plotly_framestyle

Plots._plotly_framestyle_defaults

Plots._plotly_framestyles

Plots._plotly_legend_pos

Plots._plotly_marker

Plots._plotly_markers

Plots._plotly_scale

Plots._plotly_seriestype

Plots._plotly_style

Plots._plotlyjs_attr

Plots._plotlyjs_marker

Plots._plotlyjs_scale

Plots._plotlyjs_seriestype

Plots._plotlyjs_style

Plots._positionAliases

Plots._positive_else_nan

Plots._prepare_annotations

Plots._prepare_plot_object

Plots._prepare_subplot

Plots._preprocess_args

Plots._preprocess_barlike

Plots._preprocess_binbarlike_weights

Plots._preprocess_binlike

Plots._preprocess_userrecipe

Plots._process_plotrecipe

Plots._process_seriesrecipe

Plots._process_userrecipe

Plots._process_userrecipes

Plots._pyplot_attr

Plots._pyplot_marker

Plots._pyplot_scale

Plots._pyplot_seriestype

Plots._pyplot_style

Plots._replace_linewidth

Plots._replace_markershape

Plots._savemap

Plots._scaleAliases

Plots._scale_adjusted_values

Plots._scale_base

Plots._scale_funcs

Plots._scaled_adapted_grid

Plots._series_added

Plots._series_defaults

Plots._series_index

Plots._series_recipe_deps

Plots._series_updated

Plots._shape_keys

Plots._shapes

Plots._show

Plots._showjuno

Plots._stepbins_path

Plots._styleAliases

Plots._subplot_defaults

Plots._subplot_setup

Plots._suppress_warnings

Plots._surface_like

Plots._theme

Plots._to_warn

Plots._transform_ticks

Plots._typeAliases

Plots._unicodeplots_attr

Plots._unicodeplots_marker

Plots._unicodeplots_scale

Plots._unicodeplots_seriestype

Plots._unicodeplots_style

Plots._update_axis

Plots._update_axis_colors

Plots._update_axis_links

Plots._update_clims

Plots._update_min_padding!

Plots._update_plot_args

Plots._update_plot_object

Plots._update_series_attributes!

Plots._update_subplot_args

Plots._update_subplot_colors

Plots._update_subplot_periphery

Plots._widen_seriestypes

Plots.abline! — Function.

Adds a+bx... straight line over the current plot, without changing the axis limits

Plots.adapted_grid

Plots.addExtension

Plots.addOrReplace

Plots.add_aliases

Plots.add_arrows

Plots.add_backend

Plots.add_backend_string

Plots.add_layout_pct!

Plots.add_non_underscore_aliases!

Plots.add_theme

Plots.aliases

Plots.aliasesAndAutopick

Plots.all3D

Plots.allAlphas

Plots.allFunctions

Plots.allLineTypes

Plots.allReals

Plots.allShapes

Plots.allStyles

Plots.all_seriestypes

Plots.allnan

Plots.alpha

Plots.alphacolor

RecipesBase.animate — Function.

Animate from an iterator which returns the plot args each iteration.

Plots.annotate! — Function.

annotate!(anns...)

Add annotations to an existing plot.

Arguments

anns: An AbstractVector of tuples of the form (x,y,text). The text object can be an String or PlotText

Example

julia> plot(1:10)
julia> annotate!([(7,3,"(7,3)"),(3,7,text("hey", 14, :left, :top, :green))])

Plots.annotations

Plots.anynan

Plots.arrow — Function.

arrow(args...)

Define arrowheads to apply to lines - args are style (:open or :closed), side (:head, :tail or :both), headlength and headwidth

Plots.as_gradient

Plots.attr

Plots.attr!

Plots.attributes

Plots.attrtypes

Plots.autopick

Plots.autopick_ignore_none_auto

Plots.axis_drawing_info

Plots.axis_limits

Plots.backend — Function.

Returns the current plotting package name. Initializes package on first call.

Set the plot backend.

Plots.backend_name — Function.

Returns the name of the current backend

Plots.backend_object

Plots.backend_package

Plots.backend_package_name

Plots.backends — Function.

Returns a list of supported backends

Plots.bar — Function.

bar(x,y)
bar!(x,y)

Make a bar plot of y vs x.

Arguments

bar_position: Symbol. Choose from :overlay (default), :stack. (warning: May not be implemented fully)

bar_width: nothing or Number. Width of bars in data coordinates. When nothing, chooses based on x (or y when orientation = :h).

bar_edges: Bool. Align bars to edges (true), or centers (the default)?

orientation: Symbol. Horizontal or vertical orientation for bar types. Values :h, :hor, :horizontal correspond to horizontal (sideways, anchored to y-axis), and :v, :vert, and :vertical correspond to vertical (the default).

Examples

julia> bar([1,2,3],[4,5,6],fillcolor=[:red,:green,:blue],fillalpha=[0.2,0.4,0.6])
julia> bar([(1,4),(2,5),(3,6)])

source

Plots.bar! — Function.

bar(x,y)
bar!(x,y)

Make a bar plot of y vs x.

Arguments

bar_position: Symbol. Choose from :overlay (default), :stack. (warning: May not be implemented fully)

bar_width: nothing or Number. Width of bars in data coordinates. When nothing, chooses based on x (or y when orientation = :h).

bar_edges: Bool. Align bars to edges (true), or centers (the default)?

orientation: Symbol. Horizontal or vertical orientation for bar types. Values :h, :hor, :horizontal correspond to horizontal (sideways, anchored to y-axis), and :v, :vert, and :vertical correspond to vertical (the default).

Examples

julia> bar([1,2,3],[4,5,6],fillcolor=[:red,:green,:blue],fillalpha=[0.2,0.4,0.6])
julia> bar([(1,4),(2,5),(3,6)])

source

Plots.barHack — Function.

A hacky replacement for a bar graph when the backend doesn't support bars directly. Convert it into a line chart with fillrange set.

Plots.barh

Plots.barh!

Plots.barhist — Function.

barhist(x)
barhist!(x)

Make a histogram bar plot. See histogram.

source

Plots.barhist! — Function.

barhist(x)
barhist!(x)

Make a histogram bar plot. See histogram.

source

Plots.base_color_type

Plots.base_colorant_type

Plots.bbox — Function.

bbox(x, y, w, h [,originargs...])
bbox(layout)

Create a bounding box for plotting

Plots.bbox!

Plots.bbox_to_pcts

Plots.bezier_value

Plots.binData — Function.

Make histogram-like bins of data

Plots.blue

Plots.bottom

Plots.bottompad

Plots.boxplot — Function.

boxplot(x, y)
boxplot!(x, y)

Make a box and whisker plot.

Keyword arguments

notch: Bool. Notch the box plot? (false)

range: Real. Values more than range*IQR below the first quartile or above the third quartile are shown as outliers (1.5)

outliers: Bool. Show outliers? (true)

whisker_width: Real or Symbol. Length of whiskers (:match)

Example

julia> using StatsPlots
julia> boxplot(repeat([1,2,3],outer=100),randn(300))

source

Plots.boxplot! — Function.

boxplot(x, y)
boxplot!(x, y)

Make a box and whisker plot.

Keyword arguments

notch: Bool. Notch the box plot? (false)

range: Real. Values more than range*IQR below the first quartile or above the third quartile are shown as outliers (1.5)

outliers: Bool. Show outliers? (true)

whisker_width: Real or Symbol. Length of whiskers (:match)

Example

julia> using StatsPlots
julia> boxplot(repeat([1,2,3],outer=100),randn(300))

source

Plots.brush

Plots.build_layout

Plots.buildanimation

Plots.calcMidpoints

Plots.calc_num_subplots

Plots.calc_r_extrema

Plots.ccolor

Plots.center — Function.

return the centroid of a Shape

Plots.cgrad

Plots.cgradients

Plots.cie_color_match

Plots.clamp_greys!

Plots.clibraries

Plots.clibrary

Plots.closeall — Function.

Close all open gui windows of the current backend

Plots.color

Plots.color_or_nothing!

Plots.color_type

Plots.coloralpha

Plots.colordiff

Plots.colormap

Plots.colormatch

Plots.command_idx

Plots.comp1

Plots.comp2

Plots.comp3

Plots.compute_angle

Plots.compute_gridsize

Plots.compute_x

Plots.compute_xyz

Plots.compute_y

Plots.compute_z

Plots.concatenate_fillrange

Plots.construct_categorical_data

Plots.contour — Function.

contour(x,y,z) 
contour!(x,y,z)

Draw contour lines of the Surface z.

Arguments

levels: Contour levels (if AbstractVector) or number of levels (if Integer)

fill: Bool. Fill area between contours or draw contours only (false by default)

Example

julia> x = y = range(-20, 20, length = 100)
julia> contour(x, y, (x, y) -> x^2 + y^2)

source

Plots.contour! — Function.

contour(x,y,z) 
contour!(x,y,z)

Draw contour lines of the Surface z.

Arguments

levels: Contour levels (if AbstractVector) or number of levels (if Integer)

fill: Bool. Fill area between contours or draw contours only (false by default)

Example

julia> x = y = range(-20, 20, length = 100)
julia> contour(x, y, (x, y) -> x^2 + y^2)

source

Plots.contour3d

Plots.contour3d!

Plots.contourf — Function.

An alias for contour with fill = true.

source

Plots.contourf! — Function.

An alias for contour with fill = true.

source

Plots.convertLegendValue

Plots.convertSeriesIndex

Plots.convertToAnyVector

Plots.convert_sci_unicode

Plots.convert_to_polar

Plots.coords — Function.

return the vertex points from a Shape or Segments object

Plots.createSegments

Plots.create_grid

Plots.create_grid_curly

Plots.create_grid_vcat

Plots.crop

Plots.current — Function.

current()

Returns the Plot object for the current plot

Plots.curve_points

Plots.curves — Function.

curves(x,y)
curves!(x,y)

Draw a Bezier curve from (x[1],y[1]) to (x[end],y[end]) with control points (x[2],y[2]), ..., (x[end-1],y[end]-1)

Example

julia> curves([1,2,3,4],[1,1,2,4])

source

Plots.curves! — Function.

curves(x,y)
curves!(x,y)

Draw a Bezier curve from (x[1],y[1]) to (x[end],y[end]) with control points (x[2],y[2]), ..., (x[end-1],y[end]-1)

Example

julia> curves([1,2,3,4],[1,1,2,4])

source

Plots.cvec

Plots.dateformatter

Plots.datetimeformatter

Plots.debugplots

Plots.debugshow

Plots.default — Function.

default(key) returns the current default value for that key default(key, value) sets the current default value for that key default(; kw...) will set the current default value for each key/value pair default(plotattributes, key) returns the key from plotattributes if it exists, otherwise default(key)

Plots.defaultOutputFormat

Plots.default_cgrad

Plots.default_extents

Plots.default_should_widen

Plots.defaultbox

Plots.density — Function.

density(x)
density!(x)

Make a line plot of a kernel density estimate of x.

Arguments

x: AbstractVector of samples for probability density estimation

Example

julia> using StatsPlots
julia> density(randn(100_000))

source

Plots.density! — Function.

density(x)
density!(x)

Make a line plot of a kernel density estimate of x.

Arguments

x: AbstractVector of samples for probability density estimation

Example

julia> using StatsPlots
julia> density(randn(100_000))

source

Plots.deuteranopic

Plots.directed_curve

Plots.discrete_value!

Plots.distinguishable_colors

Plots.diverging_palette

Plots.dumpcallstack

Plots.dumpdict

Plots.ensure_gradient!

Plots.eps

Plots.error_coords

Plots.error_style!

Plots.error_zipit

Plots.eval

Plots.expand_data

Plots.expand_extrema!

Plots.extendSeriesByOne

Plots.extendSeriesData

Plots.extractGroupArgs

Plots.extrema_plus_buffer

Plots.fakedata

Plots.fg_color

Plots.fg_color_sp

Plots.file_extension

Plots.filter_data

Plots.filter_data!

Plots.filter_radial_data

Plots.float_extended_type

Plots.font — Function.

font(args...)

Create a Font from a list of features. Values may be specified either as arguments (which are distinguished by type/value) or as keyword arguments.

Arguments

family: AbstractString. "serif" or "sans-serif" or "monospace"

pointsize: Integer. Size of font in points

halign: Symbol. Horizontal alignment (:hcenter, :left, or :right)

valign: Symbol. Vertical aligment (:vcenter, :top, or :bottom)

rotation: Real. Angle of rotation for text in degrees (use a non-integer type)

color: Colorant or Symbol

Examples

julia> font(8)
julia> font(family="serif",halign=:center,rotation=45.0)

Plots.frame — Function.

frame(animation[, plot])

Add a plot (the current plot if not specified) to an existing animation

Plots.from_polar

Plots.gamutmax

Plots.gamutmin

Plots.getExtension

Plots.getSeriesRGBColor

Plots.get_axis

Plots.get_clims

Plots.get_color_palette

Plots.get_fillalpha

Plots.get_fillcolor

Plots.get_linealpha

Plots.get_linecolor

Plots.get_linestyle

Plots.get_linewidth

Plots.get_markeralpha

Plots.get_markercolor

Plots.get_markerstrokealpha

Plots.get_markerstrokecolor

Plots.get_minor_ticks

Plots.get_sp_lims

Plots.get_subplot

Plots.get_subplot_index

Plots.get_ticks

Plots.get_xs

Plots.get_xy

Plots.get_ys

Plots.getattr

Plots.getplot

Plots.getxy

Plots.getxyz

Plots.gif

Plots.giffn

Plots.gr

Plots.gr_colorbar

Plots.gr_colorbar_ratio

Plots.gr_display

Plots.gr_draw_marker

Plots.gr_draw_markers

Plots.gr_fill_viewport

Plots.gr_font_family

Plots.gr_get_ticks_size

Plots.gr_getcolorind

Plots.gr_halign

Plots.gr_inqtext

Plots.gr_legend_pos

Plots.gr_lims

Plots.gr_linetype

Plots.gr_markertype

Plots.gr_nans_to_infs!

Plots.gr_plot_size

Plots.gr_polaraxes

Plots.gr_polyline

Plots.gr_set_fill

Plots.gr_set_fillcolor

Plots.gr_set_font

Plots.gr_set_gradient

Plots.gr_set_line

Plots.gr_set_linecolor

Plots.gr_set_markercolor

Plots.gr_set_output

Plots.gr_set_textcolor

Plots.gr_set_transparency

Plots.gr_set_viewport_cmap

Plots.gr_set_viewport_plotarea

Plots.gr_set_viewport_polar

Plots.gr_set_xticks_font

Plots.gr_set_yticks_font

Plots.gr_text

Plots.gr_valign

Plots.gr_view_xcenter

Plots.gr_view_ycenter

Plots.gr_viewport_from_bbox

Plots.gr_x_axislims

Plots.gr_xy_axislims

Plots.gr_y_axislims

Plots.gr_z_axislims

Plots.gray

Plots.green

Plots.grid — Function.

grid(args...; kw...)

Create a grid layout for subplots. args specify the dimensions, e.g. grid(3,2, widths = (0.6,04)) creates a grid with three rows and two columns of different width.

Plots.group_as_matrix

Plots.groupedvec2mat

Plots.gui — Function.

gui([plot])

Display a plot using the backends' gui window

Plots.guide_padding

Plots.guidefont

Plots.handleColors!

Plots.handlemissings

Plots.has_attribute_segments

Plots.has_black_border_for_default

Plots.hasattr

Plots.hascolorbar

Plots.hasgrid

Plots.hdf5

Plots.heatmap — Function.

heatmap(x,y,z)
heatmap!(x,y,z)

Plot a heatmap of the rectangular array z.

Example

julia> heatmap(randn(10,10))

source

Plots.heatmap! — Function.

heatmap(x,y,z)
heatmap!(x,y,z)

Plot a heatmap of the rectangular array z.

Example

julia> heatmap(randn(10,10))

source

Plots.heatmap_edges — Function.

create an (n+1) list of the outsides of heatmap rectangles

Plots.hex

Plots.hexbin — Function.

hexbin(x,y)
hexbin!(x,y)

Make a hexagonal binning plot (a histogram of the observations (x[i],y[i]) with hexagonal bins)

Example

julia> hexbin(randn(10_000), randn(10_000))

source

Plots.hexbin! — Function.

hexbin(x,y)
hexbin!(x,y)

Make a hexagonal binning plot (a histogram of the observations (x[i],y[i]) with hexagonal bins)

Example

julia> hexbin(randn(10_000), randn(10_000))

source

Plots.histogram — Function.

histogram(x)
histogram!(x)

Plot a histogram.

Arguments

x: AbstractVector of values to be binned

bins: Integer, NTuple{2,Integer}, AbstractVector or Symbol. Default is :auto (the Freedman-Diaconis rule). For histogram-types, defines the approximate number of bins to aim for, or the auto-binning algorithm to use (:sturges, :sqrt, :rice, :scott or :fd). For fine-grained control pass a Vector of break values, e.g. range(minimum(x), stop = maximum(x), length = 25)

weights: Vector of weights for the values in x, for weighted bin counts

normalize: Bool or Symbol. Histogram normalization mode. Possible values are: false/:none (no normalization, default), true/:pdf (normalize to a discrete Probability Density Function, where the total area of the bins is 1), :probability (bin heights sum to 1) and :density (the area of each bin, rather than the height, is equal to the counts - useful for uneven bin sizes).

bar_position: Symbol. Choose from :overlay (default), :stack. (warning: May not be implemented fully)

bar_width: nothing or Number. Width of bars in data coordinates. When nothing, chooses based on x (or y when orientation = :h).

bar_edges: Bool. Align bars to edges (true), or centers (the default)?

orientation: Symbol. Horizontal or vertical orientation for bar types. Values :h, :hor, :horizontal correspond to horizontal (sideways, anchored to y-axis), and :v, :vert, and :vertical correspond to vertical (the default).

Example

julia> histogram([1,2,1,1,4,3,8],bins=0:8)

source

Plots.histogram! — Function.

histogram(x)
histogram!(x)

Plot a histogram.

Arguments

x: AbstractVector of values to be binned

bins: Integer, NTuple{2,Integer}, AbstractVector or Symbol. Default is :auto (the Freedman-Diaconis rule). For histogram-types, defines the approximate number of bins to aim for, or the auto-binning algorithm to use (:sturges, :sqrt, :rice, :scott or :fd). For fine-grained control pass a Vector of break values, e.g. range(minimum(x), stop = maximum(x), length = 25)

weights: Vector of weights for the values in x, for weighted bin counts

normalize: Bool or Symbol. Histogram normalization mode. Possible values are: false/:none (no normalization, default), true/:pdf (normalize to a discrete Probability Density Function, where the total area of the bins is 1), :probability (bin heights sum to 1) and :density (the area of each bin, rather than the height, is equal to the counts - useful for uneven bin sizes).

bar_position: Symbol. Choose from :overlay (default), :stack. (warning: May not be implemented fully)

bar_width: nothing or Number. Width of bars in data coordinates. When nothing, chooses based on x (or y when orientation = :h).

bar_edges: Bool. Align bars to edges (true), or centers (the default)?

orientation: Symbol. Horizontal or vertical orientation for bar types. Values :h, :hor, :horizontal correspond to horizontal (sideways, anchored to y-axis), and :v, :vert, and :vertical correspond to vertical (the default).

Example

julia> histogram([1,2,1,1,4,3,8],bins=0:8)

source

Plots.histogram2d — Function.

histogram2d(x,y)
histogram2d!(x,y)

Plot a two-dimensional histogram.

Arguments

bins: Number of bins (if an Integer) or bin edges (if an AbtractVector)

weights: Vector of weights for the values in x. Each entry of x contributes its weight to the height of its bin.

Example

julia> histogram2d(randn(10_000),randn(10_000))

source

Plots.histogram2d! — Function.

histogram2d(x,y)
histogram2d!(x,y)

Plot a two-dimensional histogram.

Arguments

bins: Number of bins (if an Integer) or bin edges (if an AbtractVector)

weights: Vector of weights for the values in x. Each entry of x contributes its weight to the height of its bin.

Example

julia> histogram2d(randn(10_000),randn(10_000))

source

Plots.histogramHack — Function.

A hacky replacement for a histogram when the backend doesn't support histograms directly. Convert it into a bar chart with the appropriate x/y values.

Plots.hline — Function.

hline(y)
hline!(y)

Draw horizontal lines at positions specified by the values in the AbstractVector y

Example

julia> hline([-1,0,2])

source

Plots.hline! — Function.

hline(y)
hline!(y)

Draw horizontal lines at positions specified by the values in the AbstractVector y

Example

julia> hline([-1,0,2])

source

Plots.hspan — Function.

hspan(y)

Draw a rectangle between the horizontal line at position y[1] and the horizontal line at position y[2]. If length(y) ≥ 4, then further rectangles are drawn between y[3] and y[4], y[5] and y[6], and so on. If length(y) is odd, then the last entry of y is ignored.

Example

julia> hspan(1:6)

source

Plots.hspan! — Function.

hspan(y)

Draw a rectangle between the horizontal line at position y[1] and the horizontal line at position y[2]. If length(y) ≥ 4, then further rectangles are drawn between y[3] and y[4], y[5] and y[6], and so on. If length(y) is odd, then the last entry of y is ignored.

Example

julia> hspan(1:6)

source

Plots.html

Plots.html_body

Plots.html_head

Plots.html_to_png

Plots.hvline_limits

Plots.ignorenan_extrema — Function.

Extrema of x-values in plot

Plots.ignorenan_maximum

Plots.ignorenan_mean

Plots.ignorenan_minimum

Plots.ijulia_initialized

Plots.imageHack

Plots.inch2mm

Plots.inch2px

Plots.include

Plots.indices_and_unique_values

Plots.inline

Plots.inspectdr

Plots.invisible

Plots.invscalefunc

Plots.is3d

Plots.is_2tuple

Plots.is_attr_supported

Plots.is_marker_supported

Plots.is_scale_supported

Plots.is_seriestype_supported

Plots.is_style_supported

Plots.is_subplot_supported

Plots.isatom

Plots.iscol

Plots.isdark

Plots.isijulia

Plots.ismatrix

Plots.isnothing

Plots.isplotnull

Plots.ispolar

Plots.ispositive

Plots.isrow

Plots.isscalar

Plots.istuple

Plots.isvector

Plots.isvertical

Plots.iter_segments

Plots.js_body

Plots.labelfunc

Plots.layout_args

Plots.left

Plots.leftpad

Plots.legendEntryFromTuple

Plots.legendfont

Plots.like_histogram

Plots.like_line

Plots.like_surface

Plots.limsType

Plots.link_axes!

Plots.link_subplots

Plots.locate_annotation

Plots.lookup_aliases

Plots.make_fillrange_from_ribbon

Plots.make_fillrange_side

Plots.make_measure_hor

Plots.make_measure_vert

Plots.make_steps

Plots.makearrowhead

Plots.makecross

Plots.makekw

Plots.makeplural

Plots.makeshape — Function.

create a shape by picking points around the unit circle. n is the number of point/sides, offset is the starting angle

Plots.makestar — Function.

create a star by weaving together points from an outer and inner circle. n is the number of arms

Plots.maketuple — Function.

duplicate a single value, or pass the 2-tuple through

Plots.makevec

Plots.mapFuncOrFuncs

Plots.mapc

Plots.mapreducec

Plots.merge_with_base_supported

Plots.mm2inch

Plots.mm2px

Plots.mov

Plots.movfn

Plots.mp4

Plots.mp4fn

Plots.nanappend!

Plots.nanpush!

Plots.nansplit

Plots.nanvcat

Plots.nobigs

Plots.nop

Plots.normalize_zvals

Plots.notimpl

Plots.num_series

Plots.ohlc — Function.

ohlc(x,y::Vector{OHLC})
ohlc!(x,y::Vector{OHLC})

Make open-high-low-close plot. Each entry of y is represented by a vertical segment extending from the low value to the high value, with short horizontal segments on the left and right indicating the open and close values, respectively.

Example

julia> meanprices = cumsum(randn(100))
julia> y = OHLC[(p+rand(),p+1,p-1,p+rand()) for p in meanprices]
julia> ohlc(y)

source

Plots.ohlc! — Function.

ohlc(x,y::Vector{OHLC})
ohlc!(x,y::Vector{OHLC})

Make open-high-low-close plot. Each entry of y is represented by a vertical segment extending from the low value to the high value, with short horizontal segments on the left and right indicating the open and close values, respectively.

Example

julia> meanprices = cumsum(randn(100))
julia> y = OHLC[(p+rand(),p+1,p-1,p+rand()) for p in meanprices]
julia> ohlc(y)

source

Plots.ok

Plots.open_browser_window

Plots.optimal_ticks_and_labels

Plots.optimize_datetime_ticks

Plots.optimize_ticks

Plots.palette

Plots.parent_bbox

Plots.partialcircle — Function.

get an array of tuples of points on a circle with radius r

Plots.path3d — Function.

path3d(x,y,z)
path3d!(x,y,z)

Plot a 3D path from (x[1],y[1],z[1]) to (x[2],y[2],z[2]), ..., to (x[end],y[end],z[end]).

Example

julia> path3d([0,1,2,3],[0,1,4,9],[0,1,8,27])

source

Plots.path3d! — Function.

path3d(x,y,z)
path3d!(x,y,z)

Plot a 3D path from (x[1],y[1],z[1]) to (x[2],y[2],z[2]), ..., to (x[end],y[end],z[end]).

Example

julia> path3d([0,1,2,3],[0,1,4,9],[0,1,8,27])

source

Plots.pdf

Plots.pgfplots

Plots.pickDefaultBackend

Plots.pie — Function.

Plot a pie diagram

Plots.pie!

Plots.pie_labels

RecipesBase.plot — Function.

The main plot command. Use plot to create a new plot object, and plot! to add to an existing one:

    plot(args...; kw...)                  # creates a new plot window, and sets it to be the current
    plot!(args...; kw...)                 # adds to the `current`
    plot!(plotobj, args...; kw...)        # adds to the plot `plotobj`

There are lots of ways to pass in data, and lots of keyword arguments... just try it and it will likely work as expected. When you pass in matrices, it splits by columns. To see the list of available attributes, use the plotattr([attr]) function, where attr is the symbol :Series:, :Subplot:, :Plot or :Axis. Pass any attribute to plotattr as a String to look up its docstring; e.g. plotattr("seriestype").

Plots.plot!

Plots.plot3d — Function.

Plot with seriestype :path3d

Plots.plot3d!

Plots.plot_color

Plots.plotarea — Function.

plotarea(subplot)

Return the bounding box of a subplot

Plots.plotarea!

Plots.plotattr — Function.

plotattr([attr])

Look up the properties of a Plots attribute, or specify an attribute type. Call plotattr() for options. The information is the same as that given on https://docs.juliaplots.org/latest/attributes/.

Plots.plotly

Plots.plotly_annotation_dict

Plots.plotly_apply_aspect_ratio

Plots.plotly_axis

Plots.plotly_close_shapes

Plots.plotly_colorbar_hack

Plots.plotly_colorscale

Plots.plotly_convert_to_datetime

Plots.plotly_data

Plots.plotly_domain

Plots.plotly_font

Plots.plotly_hover!

Plots.plotly_layout

Plots.plotly_layout_json

Plots.plotly_legend_pos

Plots.plotly_link_indicies

Plots.plotly_local_file_path

Plots.plotly_native_data

Plots.plotly_polar!

Plots.plotly_polaraxis

Plots.plotly_remote_file_path

Plots.plotly_scale

Plots.plotly_series

Plots.plotly_series_json

Plots.plotly_series_segments

Plots.plotly_series_shapes

Plots.plotly_surface_data

Plots.plotlyjs

Plots.plots_heatmap

Plots.plots_heatmap!

Plots.png

Plots.portfoliocomposition

Plots.portfoliocomposition!

Plots.prepare_output

Plots.preprocessArgs! — Function.

Handle all preprocessing of args... break out colors/sizes/etc and replace aliases.

Plots.printnothing

Plots.processFillArg

Plots.processFontArg!

Plots.processGridArg!

Plots.processLineArg

Plots.processMarkerArg

Plots.processMinorGridArg!

Plots.process_annotation

Plots.process_axis_arg!

Plots.protanopic

Plots.ps

Plots.push_x!

Plots.push_y!

Plots.push_z!

Plots.px2inch

Plots.px2mm

Plots.pyplot

Plots.quiver — Function.

quiver(x,y,quiver=(u,v))
quiver!(x,y,quiver=(u,v))

Make a quiver (vector field) plot. The ith vector extends from (x[i],y[i]) to (x[i] + u[i], y[i] + v[i]).

Example

julia> quiver([1,2,3],[3,2,1],quiver=([1,1,1],[1,2,3]))

source

Plots.quiver! — Function.

quiver(x,y,quiver=(u,v))
quiver!(x,y,quiver=(u,v))

Make a quiver (vector field) plot. The ith vector extends from (x[i],y[i]) to (x[i] + u[i], y[i] + v[i]).

Example

julia> quiver([1,2,3],[3,2,1],quiver=([1,1,1],[1,2,3]))

source

Plots.quiver_using_arrows

Plots.quiver_using_hack

Plots.recompute_lengths

Plots.red

Plots.reducec

Plots.regressionXY

Plots.replaceAlias!

Plots.replaceAliases!

Plots.replaceType

Plots.replace_image_with_heatmap

Plots.reset_defaults

Plots.reset_extrema!

Plots.resolve_mixed

Plots.rgb_string

Plots.rgba_string

Plots.right

Plots.rightpad

Plots.rotate — Function.

rotate an object in space

Plots.rotate!

Plots.rotate_x

Plots.rotate_y

Plots.round_limits

Plots.rowsize

Plots.savefig — Function.

savefig([plot,] filename)

Save a Plot (the current plot if plot is not passed) to file. The file type is inferred from the file extension. All backends support png and pdf file types, some also support svg, ps, eps, html and tex.

Plots.scale

Plots.scale!

Plots.scalefontsize

Plots.scalefontsizes — Function.

scalefontsizes(factor::Number)

Scales all current font sizes by factor. For example scalefontsizes(1.1) increases all current font sizes by 10%. To reset to initial sizes, use scalefontsizes()

scalefontsizes()

Resets font sizes to initial default values.

Plots.scalefunc

Plots.scatter — Function.

scatter(x,y)
scatter!(x,y)

Make a scatter plot of y vs x.

Examples

julia> scatter([1,2,3],[4,5,6],markersize=[3,4,5],markercolor=[:red,:green,:blue])
julia> scatter([(1,4),(2,5),(3,6)])

source

Plots.scatter! — Function.

scatter(x,y)
scatter!(x,y)

Make a scatter plot of y vs x.

Examples

julia> scatter([1,2,3],[4,5,6],markersize=[3,4,5],markercolor=[:red,:green,:blue])
julia> scatter([(1,4),(2,5),(3,6)])

source

Plots.scatter3d — Function.

scatter3d(x,y,z)
scatter3d!(x,y,z)

Make a 3D scatter plot.

Example

julia> scatter3d([0,1,2,3],[0,1,4,9],[0,1,8,27])

source

Plots.scatter3d! — Function.

scatter3d(x,y,z)
scatter3d!(x,y,z)

Make a 3D scatter plot.

Example

julia> scatter3d([0,1,2,3],[0,1,4,9],[0,1,8,27])

source

Plots.scatterhist — Function.

scatterhist(x)
scatterhist!(x)

Make a histogram scatter plot (bin counts are represented using points instead of bars). See histogram.

source

Plots.scatterhist! — Function.

scatterhist(x)
scatterhist!(x)

Make a histogram scatter plot (bin counts are represented using points instead of bars). See histogram.

source

Plots.sequential_palette

Plots.series_annotations

Plots.series_annotations_shapes!

Plots.series_list

Plots.series_recipe_dependencies

Plots.seriestype_supported

Plots.set_theme

Plots.setxy!

Plots.setxyz!

Plots.shape_coords

Plots.shape_data

Plots.should_add_to_legend

Plots.show_png_from_html

Plots.showaxis

Plots.showgradient

Plots.showgradient!

Plots.showjuno

Plots.showlibrary

Plots.showlibrary!

Plots.showtheme

Plots.showtheme!

Plots.shrink_by

Plots.slice_arg

Plots.slice_arg!

Plots.sortedkeys

Plots.split_kw

Plots.splittable_kw

Plots.spy

Plots.spy!

Plots.standalone_html

Plots.standalone_html_window

Plots.stephist — Function.

stephist(x)
stephist(x)

Make a histogram step plot (bin counts are represented using horizontal lines instead of bars). See histogram.

source

Plots.stephist! — Function.

stephist(x)
stephist(x)

Make a histogram step plot (bin counts are represented using horizontal lines instead of bars). See histogram.

source

Plots.sticks — Function.

sticks(x,y)
sticks!(x,y)

Draw a stick plot of y vs x.

Example

julia> sticks(1:10) 

source

Plots.sticks! — Function.

sticks(x,y)
sticks!(x,y)

Draw a stick plot of y vs x.

Example

julia> sticks(1:10) 

source

Plots.sticksHack — Function.

A hacky replacement for a sticks graph when the backend doesn't support sticks directly. Convert it into a line chart that traces the sticks, and a scatter that sets markers at the points.

Plots.straightline_data

Plots.stroke — Function.

stroke(args...; alpha = nothing)

Define the properties of the stroke used in plotting lines

Plots.supported_attrs

Plots.supported_markers

Plots.supported_scales

Plots.supported_seriestypes

Plots.supported_styles

Plots.surface — Function.

surface(x,y,z)
surface!(x,y,z)

Draw a 3D surface plot.

Example

julia> x = y = range(-3, 3, length = 100)
julia> surface(x, y, (x, y) -> sinc(norm([x, y])))

source

Plots.surface! — Function.

surface(x,y,z)
surface!(x,y,z)

Draw a 3D surface plot.

Example

julia> x = y = range(-3, 3, length = 100)
julia> surface(x, y, (x, y) -> sinc(norm([x, y])))

source

Plots.svg

Plots.test_examples — Function.

test_examples(pkgname[, idx]; debug = false, disp = true, sleep = nothing, skip = [], only = nothing

Run the idx test example for a given backend, or all examples if idx is not specified.

Plots.tex

Plots.text — Function.

text(string, args...; kw...)

Create a PlotText object wrapping a string with font info, for plot annotations. args and kw are passed to font.

Plots.text_size — Function.

Returns the (width,height) of a text label.

Plots.theme — Function.

theme(s::Symbol)

Specify the colour theme for plots.

Plots.tick_padding

Plots.tickfont

Plots.ticksType

Plots.timeformatter

Plots.title! — Function.

Add title to an existing plot

Plots.title_padding

Plots.titlefont

Plots.to_nan

Plots.to_pixels

Plots.top

Plots.toppad

Plots.tovec

Plots.translate

Plots.translate! — Function.

translate a Shape in space

Plots.transpose_z

Plots.tritanopic

Plots.trueOrAllTrue

Plots.tryrange

Plots.twinx — Function.

Adds a new, empty subplot overlayed on top of sp, with a mirrored y-axis and linked x-axis.

Plots.unicodeplots

Plots.unzip

Plots.update_child_bboxes!

Plots.update_inset_bboxes!

Plots.update_position!

Plots.use_local_dependencies

Plots.vertices

Plots.viewport_plotarea

Plots.violin — Function.

violin(x,y,z)
violin!(x,y,z)

Make a violin plot.

Example

julia> violin(repeat([1,2,3],outer=100),randn(300))

source

Plots.violin! — Function.

violin(x,y,z)
violin!(x,y,z)

Make a violin plot.

Example

julia> violin(repeat([1,2,3],outer=100),randn(300))

source

Plots.vline — Function.

vline(x)
vline!(x)

Draw vertical lines at positions specified by the values in the AbstractVector x

Example

julia> vline([-1,0,2])

source

Plots.vline! — Function.

vline(x)
vline!(x)

Draw vertical lines at positions specified by the values in the AbstractVector x

Example

julia> vline([-1,0,2])

source

Plots.vspan — Function.

vspan(x)

Draw a rectangle between the vertical line at position x[1] and the vertical line at position x[2]. If length(x) ≥ 4, then further rectangles are drawn between x[3] and x[4], x[5] and x[6], and so on. If length(x) is odd, then the last entry of x is ignored.

Example

julia> vspan(1:6)

source

Plots.vspan! — Function.

vspan(x)

Draw a rectangle between the vertical line at position x[1] and the vertical line at position x[2]. If length(x) ≥ 4, then further rectangles are drawn between x[3] and x[4], x[5] and x[6], and so on. If length(x) is odd, then the last entry of x is ignored.

Example

julia> vspan(1:6)

source

Plots.wand_edges

Plots.warnOnUnsupported

Plots.warnOnUnsupported_args

Plots.warnOnUnsupported_scales

Plots.warn_on_deprecated_backend

Plots.weave — Function.

interleave 2 vectors into each other (like a zipper's teeth)

Plots.weighted_color_mean

Plots.whitebalance

Plots.widen

Plots.wireframe — Function.

wireframe(x,y,z)
wireframe!(x,y,z)

Draw a 3D wireframe plot.

Example

julia> wireframe(1:10,1:10,randn(10,10))

source

Plots.wireframe! — Function.

wireframe(x,y,z)
wireframe!(x,y,z)

Draw a 3D wireframe plot.

Example

julia> wireframe(1:10,1:10,randn(10,10))

source

Plots.with — Function.

Allows temporary setting of backend and defaults for Plots. Settings apply only for the do block. Example:

with(:gr, size=(400,400), type=:histogram) do
  plot(rand(10))
  plot(rand(10))
end

Plots.wrap

Plots.wrap_surfaces

Plots.wraptuple

Plots.write_temp_html

Plots.xaxis! — Function.

Specify x axis attributes for an existing plot

Plots.xflip! — Function.

Flip the current plots' x axis

Plots.xgrid!

Plots.xlabel! — Function.

Add xlabel to an existing plot

Plots.xlims — Function.

xlims([plt])

Returns the x axis limits of the current plot or subplot

Plots.xlims! — Function.

Set xlims for an existing plot

Plots.xmax — Function.

Largest x in plot

Plots.xmin — Function.

Smallest x in plot

Plots.xticks! — Function.

Set xticks for an existing plot

Plots.xyY

Plots.xyYA

Plots.xy_mm_to_pcts

Plots.yaxis! — Function.

Specify x axis attributes for an existing plot

Plots.yflip! — Function.

Flip the current plots' y axis

Plots.ygrid!

Plots.ylabel! — Function.

Add ylabel to an existing plot

Plots.ylims — Function.

ylims([plt])

Returns the y axis limits of the current plot or subplot

Plots.ylims! — Function.

Set ylims for an existing plot

Plots.yticks! — Function.

Set yticks for an existing plot

Plots.zlims — Function.

zlims([plt])

Returns the z axis limits of the current plot or subplot

Plots.zlims! — Function.

Set zlims for an existing plot

Plots.zvalues