Docstrings

JSON.Common — Module.

Internal implementation detail.

JSON.JSON

JSON.Writer.JSONText — Type.

JSONText(s::AbstractString)

JSONText is a wrapper around a Julia string representing JSON-formatted text, which is inserted as-is in the JSON output of JSON.print and JSON.json for compact output, and is otherwise re-parsed for pretty-printed output.

s must contain valid JSON text. Otherwise compact output will contain the malformed s and other serialization output will throw a parsing exception.

JSON.Parser

JSON.Serializations — Module.

JSON writer serialization contexts.

This module defines the Serialization abstract type and several concrete implementations, as they relate to JSON.

JSON.Writer

JSON.eval

JSON.include

JSON.json