Docstrings

JuMP.@Expression

JuMP.@LinearConstraint

JuMP.@LinearConstraints

JuMP.@NLconstraint

JuMP.@NLconstraints

JuMP.@NLexpression

JuMP.@NLexpressions

JuMP.@NLobjective

JuMP.@NLparameter

JuMP.@QuadConstraint

JuMP.@QuadConstraints

JuMP.@SDconstraint

JuMP.@SDconstraints

JuMP.@SOCConstraint

JuMP.@SOCConstraints

JuMP.@addConstraint

JuMP.@addConstraints

JuMP.@addLazyConstraint

JuMP.@addNLConstraint

JuMP.@addNLConstraints

JuMP.@addSDPConstraint

JuMP.@addusercut

JuMP.@constraint

JuMP.@constraintref

JuMP.@constraints

JuMP.@defConstrRef

JuMP.@defExpr

JuMP.@defNLExpr

JuMP.@defNLParam

JuMP.@defVar

JuMP.@defVars

JuMP.@deprecate_macro

JuMP.@expression

JuMP.@expressions

JuMP.@lazyconstraint

JuMP.@objective

JuMP.@processNLExpr

JuMP.@setNLObjective

JuMP.@setObjective

JuMP.@usercut

JuMP.@variable

JuMP.@variables

JuMP.AbstractConstraint

JuMP.AbstractJuMPScalar

JuMP.AbstractModel

JuMP.AffExpr

JuMP.CallbackAbort

JuMP.ConstraintRef

JuMP.CutCallback

JuMP.DIMS

JuMP.DisableNLPResolve

JuMP.EMPTYSTRING

JuMP.EnableNLPResolve

JuMP.FunctionStorage

JuMP.GenericAffExpr

JuMP.GenericNorm

JuMP.GenericNormExpr

JuMP.GenericQuadConstraint

JuMP.GenericQuadExpr

JuMP.GenericRangeConstraint

JuMP.GenericSOCConstraint

JuMP.GenericSOCExpr

JuMP.HeuristicCallback

JuMP.IJuliaMode

JuMP.IndexPair

JuMP.IndexedVector

JuMP.InfoCallback

JuMP.JuMP

JuMP.JuMPArray

JuMP.JuMPCallback

JuMP.JuMPContainer

JuMP.JuMPContainerData

JuMP.JuMPContainer_from

JuMP.JuMPDict

JuMP.JuMPNLPEvaluator

JuMP.JuMPScalars

JuMP.JuMPTypes

JuMP.KeyIterator

JuMP.LazyCallback

JuMP.LinConstrRef

JuMP.LinearConstraint

JuMP.LinearTermIterator

JuMP.Model

JuMP.NLPData

JuMP.NLPEvaluator

JuMP.NonlinearConstraint

JuMP.NonlinearExprData

JuMP.NonlinearExpression

JuMP.NonlinearParameter

JuMP.Norm

JuMP.PRINT_ZERO_TOL

JuMP.PSDCone

JuMP.ParameterPrintWrapper

JuMP.PrintMode

JuMP.PrintSymbols

JuMP.ProblemTraits

JuMP.QuadConstraint

JuMP.QuadExpr

JuMP.REPLMode

JuMP.SDConstraint

JuMP.SOCConstraint

JuMP.SOCExpr

JuMP.SOSConstraint

JuMP.StopTheSolver

JuMP.SubexpressionPrintWrapper

JuMP.SubexpressionStorage

JuMP.UnsetSolver

JuMP.UserAutoDiffEvaluator

JuMP.UserFunctionEvaluator

JuMP.ValueIterator

JuMP.Variable

JuMP.VariableNotOwnedError

JuMP.VariablePrintWrapper

JuMP.VectorView

JuMP._NLExpr

JuMP.__next

JuMP._add_zero

JuMP._buildInternalModel_nlp

JuMP._build_norm

JuMP._canonicalize_sense

JuMP._construct_constraint!

JuMP._dot

JuMP._fillwithzeros

JuMP._gendict_checkgen

JuMP._getDual

JuMP._getValue

JuMP._getmodel

JuMP._hesslag_structure

JuMP._ind2sub

JuMP._innercontainer

JuMP._isless

JuMP._localvar

JuMP._map

JuMP._mapInner

JuMP._matmul

JuMP._matmult

JuMP._matprod_type

JuMP._multiply!

JuMP._multiplyt!

JuMP._next

JuMP._nlexprerr

JuMP._print_array

JuMP._return_array

JuMP._return_arrayt

JuMP._similar

JuMP._size

JuMP._sizehint_expr!

JuMP._throw_transpose_error

JuMP._to_cartesian

JuMP._to_cartesian_range_sub

JuMP._unspecifiedstate

JuMP._values

JuMP._vecaff

JuMP._vectorize_like

JuMP._warnnan

JuMP.addConstraint

JuMP.addCutCallback

JuMP.addHeuristicCallback

JuMP.addInfoCallback

JuMP.addLazyCallback

JuMP.addLazyConstraint

JuMP.addNLconstraint

JuMP.addQuadratics

JuMP.addSOS

JuMP.addSOS1

JuMP.addSOS2

JuMP.addSolution

JuMP.addToExpression

JuMP.addUserCut

JuMP.addVectorizedConstraint

JuMP.add_JuMP_prefix

JuMP.addconstraint

JuMP.addcutcallback

JuMP.addelt!

JuMP.addheuristiccallback

JuMP.addinfocallback

JuMP.addkwargs!

JuMP.addlazycallback

JuMP.addlazyconstraint

JuMP.addsolution

JuMP.addtoexpr

JuMP.addtoexpr_reorder

JuMP.addusercut

JuMP.affToExpr

JuMP.affToStr

JuMP.aff_str

JuMP.assert_isfinite

JuMP.assert_validmodel

JuMP.attach_callbacks

JuMP.build

JuMP.buildInternalModel

JuMP.buildrefsets

JuMP.chgConstrRHS

JuMP.coeftype

JuMP.collect_expr!

JuMP.coloncheck

JuMP.conToStr

JuMP.con_str

JuMP.conicdata

JuMP.constraint_error

JuMP.constraintbounds

JuMP.constructSOS

JuMP.constructconstraint!

JuMP.constructvariable!

JuMP.cont_str

JuMP.cont_str_set

JuMP.curly_to_generator

JuMP.cutcallback

JuMP.dependson

JuMP.esc_nonconstant

JuMP.eval

JuMP.exprToStr

JuMP.expr_str

JuMP.exprstr

JuMP.fillConicDuals

JuMP.fillConicRedCosts

JuMP.fill_bounds_constr!

JuMP.fill_var_names

JuMP.fillconstr!

JuMP.fillconstrLHS!

JuMP.fillconstrRHS!

JuMP.fillconstrtorow!

JuMP.fix

JuMP.forward_eval_all

JuMP.gendict

JuMP.getCategory

JuMP.getConstraintBounds

JuMP.getDual

JuMP.getInternalModel

JuMP.getLinearIndex

JuMP.getLower

JuMP.getName

JuMP.getNumBndRows

JuMP.getNumRows

JuMP.getNumSDPRows

JuMP.getNumSOCRows

JuMP.getNumSymRows

JuMP.getObjective

JuMP.getObjectiveSense

JuMP.getObjectiveValue

JuMP.getSDrowsinfo

JuMP.getUpper

JuMP.getValue

JuMP.getVar

JuMP.getbarrieriter

JuMP.getcategory

JuMP.getconicdualaux

JuMP.getconstraint

JuMP.getdual

JuMP.getdualwarn

JuMP.getloopedcode

JuMP.getlowerbound

JuMP.getmeta

JuMP.getname

JuMP.getnodecount

JuMP.getobjbound

JuMP.getobjective

JuMP.getobjectivebound

JuMP.getobjectivesense

JuMP.getobjectivevalue

JuMP.getobjgap

JuMP.getrawsolver

JuMP.getsimplexiter

JuMP.getsolvetime

JuMP.getupperbound

JuMP.getvalue

JuMP.getvalue_warn

JuMP.getvaluewarn

JuMP.getvariable

JuMP.hasdependentsets

JuMP.hasmeta

JuMP.hessian_slice

JuMP.hessian_slice_inner

JuMP.heurcallback

JuMP.ijulia

JuMP.include

JuMP.indexability

JuMP.infocallback

JuMP.initNLP

JuMP.internalmodel

JuMP.is_complex_expr

JuMP.isdependent

JuMP.isprod

JuMP.issum

JuMP.lazycallback

JuMP.let_code_block

JuMP.linearindex

JuMP.linearterms

JuMP.localvar

JuMP.macro_assign_and_return — Function.

macro_assign_and_return(code, variable, name;
                        register_fun::Union{Nothing, Function}=nothing,
                        model=nothing)

Return runs code in a local scope which returns the value of variable and then assign variable to name. If register_fun is given, register_fun(model, name, variable) is called.

JuMP.macro_return — Function.

macro_return(model, code, variable)

Return a block of code that 1. runs the code block code in a local scope and 2. returns the value of a local variable named variable. variable must reference a variable defined by code.

JuMP.mapcontainer_warn

JuMP.math

JuMP.merge_duplicates

JuMP.metadata

JuMP.model_str

JuMP.mul!

JuMP.newparameter

JuMP.no_solver_error

JuMP.norm_str

JuMP.notindexable_start

JuMP.numnlconstr

JuMP.numsdconstr

JuMP.numsocconstr

JuMP.numsosconstr

JuMP.offdiagsdpvars

JuMP.one_indexed

JuMP.op_hint

JuMP.operator_warn

JuMP.parseCurly

JuMP.parseExpr

JuMP.parseExprToplevel

JuMP.parseGenerator

JuMP.parseGeneratorNorm

JuMP.parseGeneratorSum

JuMP.parseIdxSet

JuMP.parseNLExpr

JuMP.parseNLExpr_runtime

JuMP.parseNorm

JuMP.parseSum

JuMP.parse_conditions

JuMP.parsegen

JuMP.prepAffObjective

JuMP.prepConstrBounds

JuMP.prepConstrMatrix

JuMP.print_shortest

JuMP.printdata

JuMP.processNLExpr

JuMP.pushmeta!

JuMP.quadToExpr

JuMP.quadToStr

JuMP.quad_str

JuMP.register

JuMP.registerNLFunction

JuMP.registercallbacks

JuMP.registercon

JuMP.registerobject

JuMP.registervar

JuMP.reinterpret_unsafe

JuMP.repl

JuMP.rescaleSDcols!

JuMP.reverse_eval_all

JuMP.rhs

JuMP.rmz!

JuMP.sense

JuMP.sensemap

JuMP.setCategory

JuMP.setLower

JuMP.setNLobjective

JuMP.setName

JuMP.setObjective

JuMP.setObjectiveSense

JuMP.setPrintHook

JuMP.setRHS

JuMP.setSolutionValue!

JuMP.setSolveHook

JuMP.setSolver

JuMP.setUpper

JuMP.setValue

JuMP.setcategory

JuMP.setlowerbound

JuMP.setname

JuMP.setobjective

JuMP.setobjectivesense

JuMP.setprinthook

JuMP.setsolutionvalue

JuMP.setsolvehook

JuMP.setsolver

JuMP.setupperbound

JuMP.setvalue

JuMP.simplify_expression

JuMP.size

JuMP.solve

JuMP.solvenlp

JuMP.spliceref

JuMP.storecontainerdata

JuMP.str_round

JuMP.suggestedsolvers

JuMP.tapeToExpr

JuMP.tidy_warmstart

JuMP.tryParseIdxSet

JuMP.val_str

JuMP.validmodel

JuMP.var_cats

JuMP.var_str

JuMP.variable_error

JuMP.variable_range_to_cone!

JuMP.variabletype

JuMP.varname_generic

JuMP.varname_given

JuMP.vartypes_without_fixed

JuMP.verify_ownership

JuMP.warn_curly

JuMP.warn_once

JuMP.writeLP

JuMP.writeMPS