Docstrings

Optim.@add_linesearch_fields

Optim.@brenttrace

Optim.@def

Optim.@goldensectiontrace

Optim.@initial_linesearch

Optim.@pack_SAMIN

Optim.@pack_SAMIN!

Optim.@unpack_SAMIN

Optim.AbstractBarrierState

Optim.AbstractConstrainedOptimizer

Optim.AbstractNGMRES

Optim.AbstractOptimizer

Optim.AbstractOptimizerState

Optim.AcceleratedGradientDescent

Optim.AcceleratedGradientDescentState

Optim.AdaptiveParameters

Optim.AffineSimplexer

Optim.BFGS — Type.

BFGS

Constructor

BFGS(; alphaguess = LineSearches.InitialStatic(),
       linesearch = LineSearches.HagerZhang(),
       initial_invH = x -> Matrix{eltype(x)}(I, length(x), length(x)),
       manifold = Flat())

Description

The BFGS method implements the Broyden-Fletcher-Goldfarb-Shanno algorithm as described in Nocedal and Wright (sec. 8.1, 1999) and the four individual papers Broyden (1970), Fletcher (1970), Goldfarb (1970), and Shanno (1970). It is a quasi-Newton method that updates an approximation to the Hessian using past approximations as well as the gradient. See also the limited memory variant LBFGS for an algorithm that is more suitable for high dimensional problems.

References

Wright, S. J. and J. Nocedal (1999), Numerical optimization. Springer Science 35.67-68: 7.

Broyden, C. G. (1970), The convergence of a class of double-rank minimization algorithms, Journal of the Institute of Mathematics and Its Applications, 6: 76–90.

Fletcher, R. (1970), A New Approach to Variable Metric Algorithms, Computer Journal, 13 (3): 317–322,

Goldfarb, D. (1970), A Family of Variable Metric Updates Derived by Variational Means, Mathematics of Computation, 24 (109): 23–26,

Shanno, D. F. (1970), Conditioning of quasi-Newton methods for function minimization, Mathematics of Computation, 24 (111): 647–656.

Optim.BFGSState

Optim.BarrierLineSearch — Type.

BarrierLineSearch{T}

Parameters for interior-point line search methods that use only the value

Optim.BarrierLineSearchGrad — Type.

BarrierLineSearchGrad{T}

Parameters for interior-point line search methods that exploit the slope.

Optim.BarrierStateVars

Optim.Brent — Type.

Brent

Constructor

    Brent(;)

Description

Also known as the Brent-Dekker algorith, Brent is a univariate optimization algorithm for minimizing functions on some interval [a,b]. The method uses bisection to find a zero of the gradient. If the original interval contains a minimum, bisection will reliably find the solution, but can be slow. To this end Brent combines bisection with the secant method and inverse quadratic interpolation to accelerate convergence.

References

R. P. Brent (2002) Algorithms for Minimization Without Derivatives. Dover edition.

Optim.ConjugateGradient — Type.

Conjugate Gradient Descent

Constructor

ConjugateGradient(; alphaguess = LineSearches.InitialHagerZhang(),
linesearch = LineSearches.HagerZhang(),
eta = 0.4,
P = nothing,
precondprep = (P, x) -> nothing,
manifold = Flat())

The strictly positive constant $eta$ is used in determining the next step direction, and the default here deviates from the one used in the original paper (where it was $0.01$). See more details in the original papers referenced below.

Description

The ConjugateGradient method implements Hager and Zhang (2006) and elements from Hager and Zhang (2013). Notice, the default linesearch is HagerZhang from LineSearches.jl. This line search is exactly the one proposed in Hager and Zhang (2006).

References

W. W. Hager and H. Zhang (2006) Algorithm 851: CG_DESCENT, a conjugate gradient method with guaranteed descent. ACM Transactions on Mathematical Software 32: 113-137.

W. W. Hager and H. Zhang (2013), The Limited Memory Conjugate Gradient Method. SIAM Journal on Optimization, 23, pp. 2150-2168.

Optim.ConjugateGradientState

Optim.ConstrainedOptimizer

Optim.FirstOrderOptimizer

Optim.FixedParameters

Optim.Flat — Type.

Flat Euclidean space {R,C}^N, with projections equal to the identity.

Optim.Fminbox — Type.

Fminbox

Constructor

Fminbox(method::T,
        mu0::Tf
        mufactor::Tf
        precondprep::P)

Optim.GoldenSection — Type.

GoldenSection

Constructor

    GoldenSection(;)

Description

The GoldenSection method seeks to minimize a univariate function on an interval [a, b]. At all times the algorithm maintains a tuple of three minimizer candidates (c, d, e) where $c<d<e$ such that the ratio of the largest to the smallest interval is the Golden Ratio.

References

https://en.wikipedia.org/wiki/Golden-section_search

Optim.GradientDescent — Type.

Gradient Descent

Constructor

GradientDescent(; alphaguess = LineSearches.InitialHagerZhang(),
linesearch = LineSearches.HagerZhang(),
P = nothing,
precondprep = (P, x) -> nothing)

Keywords are used to control choice of line search, and preconditioning.

Description

The GradientDescent method a simple gradient descent algorithm, that is the search direction is simply the negative gradient at the current iterate, and then a line search step is used to compute the final step. See Nocedal and Wright (ch. 2.2, 1999) for an explanation of the approach.

References

Nocedal, J. and Wright, S. J. (1999), Numerical optimization. Springer Science 35.67-68: 7.

Optim.GradientDescentState

Optim.Hf

Optim.IPNewton — Type.

Interior-point Newton

Constructor

IPNewton(; linesearch::Function = Optim.backtrack_constrained_grad,
         μ0::Union{Symbol,Number} = :auto,
         show_linesearch::Bool = false)

The initial barrier penalty coefficient μ0 can be chosen as a number, or set to :auto to let the algorithm decide its value, see initialize_μ_λ!.

Note: For constrained optimization problems, we recommend always enabling allow_f_increases and successive_f_tol in the options passed to optimize. The default is set to Optim.Options(allow_f_increases = true, successive_f_tol = 2).

As of February 2018, the line search algorithm is specialised for constrained interior-point methods. In future we hope to support more algorithms from LineSearches.jl.

Description

The IPNewton method implements an interior-point primal-dual Newton algorithm for solving nonlinear, constrained optimization problems. See Nocedal and Wright (Ch. 19, 2006) for a discussion of interior-point methods for constrained optimization.

References

The algorithm was originally written by Tim Holy (@timholy, tim.holy@gmail.com).

J Nocedal, SJ Wright (2006), Numerical optimization, second edition. Springer.

A Wächter, LT Biegler (2006), On the implementation of an interior-point filter line-search algorithm for large-scale nonlinear programming. Mathematical Programming 106 (1), 25-57.

Optim.IPNewtonState

Optim.IPOptimizer

Optim.InverseDiagonal

Optim.KrylovTrustRegion

Optim.KrylovTrustRegionState

Optim.LBFGS — Type.

LBFGS

Constructor

LBFGS(; m::Integer = 10,
alphaguess = LineSearches.InitialStatic(),
linesearch = LineSearches.HagerZhang(),
P=nothing,
precondprep = (P, x) -> nothing,
manifold = Flat(),
scaleinvH0::Bool = true && (typeof(P) <: Nothing))

LBFGS has two special keywords; the memory length m, and the scaleinvH0 flag. The memory length determines how many previous Hessian approximations to store. When scaleinvH0 == true, then the initial guess in the two-loop recursion to approximate the inverse Hessian is the scaled identity, as can be found in Nocedal and Wright (2nd edition) (sec. 7.2).

In addition, LBFGS supports preconditioning via the P and precondprep keywords.

Description

The LBFGS method implements the limited-memory BFGS algorithm as described in Nocedal and Wright (sec. 7.2, 2006) and original paper by Liu & Nocedal (1989). It is a quasi-Newton method that updates an approximation to the Hessian using past approximations as well as the gradient.

References

Wright, S. J. and J. Nocedal (2006), Numerical optimization, 2nd edition. Springer

Liu, D. C. and Nocedal, J. (1989). "On the Limited Memory Method for Large Scale Optimization". Mathematical Programming B. 45 (3): 503–528

Optim.LBFGSState

Optim.Manifold

Optim.ManifoldObjective

Optim.MaximizationWrapper

Optim.MomentumGradientDescent

Optim.MomentumGradientDescentState

Optim.MultivariateOptimizationResults

Optim.NGMRES — Type.

N-GMRES

Constructor

NGMRES(;
        alphaguess = LineSearches.InitialStatic(),
        linesearch = LineSearches.HagerZhang(),
        manifold = Flat(),
        wmax::Int = 10,
        ϵ0 = 1e-12,
        nlprecon = GradientDescent(
            alphaguess = LineSearches.InitialStatic(alpha=1e-4,scaled=true),
            linesearch = LineSearches.Static(),
            manifold = manifold),
        nlpreconopts = Options(iterations = 1, allow_f_increases = true),
      )

Description

This algorithm takes a step given by the nonlinear preconditioner nlprecon and proposes an accelerated step by minimizing an approximation of the (ll_2) residual of the gradient on a subspace spanned by the previous wmax iterates.

N-GMRES was originally developed for solving nonlinear systems [1], and reduces to GMRES for linear problems. Application of the algorithm to optimization is covered, for example, in [2].

References

[1] De Sterck. Steepest descent preconditioning for nonlinear GMRES optimization. NLAA, 2013. [2] Washio and Oosterlee. Krylov subspace acceleration for nonlinear multigrid schemes. ETNA, 1997.

Optim.NGMRESState

Optim.NMParameters

Optim.NelderMead — Type.

NelderMead

Constructor

NelderMead(; parameters = AdaptiveParameters(),
             initial_simplex = AffineSimplexer())

The constructor takes 2 keywords:

parameters, an instance of either AdaptiveParameters or FixedParameters,

and is used to generate parameters for the Nelder-Mead Algorithm

initial_simplex, an instance of AffineSimplexer

Description

Our current implementation of the Nelder-Mead algorithm is based on [1] and [3]. Gradient-free methods can be a bit sensitive to starting values and tuning parameters, so it is a good idea to be careful with the defaults provided in Optim.jl.

Instead of using gradient information, Nelder-Mead is a direct search method. It keeps track of the function value at a number of points in the search space. Together, the points form a simplex. Given a simplex, we can perform one of four actions: reflect, expand, contract, or shrink. Basically, the goal is to iteratively replace the worst point with a better point. More information can be found in [1], [2] or [3].

References

[1] Nelder, John A. and R. Mead (1965). "A simplex method for function minimization". Computer Journal 7: 308–313. doi:10.1093/comjnl/7.4.308

[2] Lagarias, Jeffrey C., et al. "Convergence properties of the Nelder–Mead simplex method in low dimensions." SIAM Journal on Optimization 9.1 (1998): 112-147

[3] Gao, Fuchang and Lixing Han (2010). "Implementing the Nelder-Mead simplex algorithm with adaptive parameters". Computational Optimization and Applications. doi:10.1007/s10589-010-9329-3

Optim.NelderMeadState

Optim.Newton — Type.

Newton

Constructor

Newton(; alphaguess = LineSearches.InitialStatic(),
linesearch = LineSearches.HagerZhang())

Description

The Newton method implements Newton's method for optimizing a function. We use a special factorization from the package PositiveFactorizations.jl to ensure that each search direction is a direction of descent. See Wright and Nocedal and Wright (ch. 6, 1999) for a discussion of Newton's method in practice.

References

Nocedal, J. and S. J. Wright (1999), Numerical optimization. Springer Science 35.67-68: 7.

Optim.NewtonState

Optim.NewtonTrustRegion — Type.

NewtonTrustRegion

Constructor

NewtonTrustRegion(; initial_delta = 1.0,
                    delta_hat = 100.0,
                    eta = 0.1,
                    rho_lower = 0.25,
                    rho_upper = 0.75)

The constructor has 5 keywords:

initial_delta, the starting trust region radius

delta_hat, the largest allowable trust region radius

eta, when rho is at least eta, accept the step

rho_lower, when rho is less than rho_lower, shrink the trust region

rho_upper, when rho is greater than rho_upper, grow the trust region

Description

The NewtonTrustRegion method implements Newton's method with a trust region for optimizing a function. The method is designed to take advantage of the second-order information in a function's Hessian, but with more stability that Newton's method when functions are not globally well-approximated by a quadratic. This is achieved by repeatedly minimizing quadratic approximations within a dynamically-sized trust region in which the function is assumed to be locally quadratic. See Wright and Nocedal and Wright (ch. 4, 1999) for a discussion of trust-region methods in practice.

References

Nocedal, J. and S. J. Wright (1999), Numerical optimization. Springer Science 35.67-68: 7.

Optim.NewtonTrustRegionState

Optim.NonDifferentiable

Optim.OACCEL — Type.

O-ACCEL

Constructor

OACCEL(;manifold::Manifold = Flat(),
       alphaguess = LineSearches.InitialStatic(),
       linesearch = LineSearches.HagerZhang(),
       nlprecon = GradientDescent(
           alphaguess = LineSearches.InitialStatic(alpha=1e-4,scaled=true),
           linesearch = LineSearches.Static(),
           manifold = manifold),
       nlpreconopts = Options(iterations = 1, allow_f_increases = true),
       ϵ0 = 1e-12,
       wmax::Int = 10)

Description

This algorithm takes a step given by the nonlinear preconditioner nlprecon and proposes an accelerated step by minimizing an approximation of the objective on a subspace spanned by the previous wmax iterates.

O-ACCEL is a slight tweak of N-GMRES, first presented in [1].

References

[1] Riseth. Objective acceleration for unconstrained optimization. 2018.

Optim.OnceDifferentiable

Optim.Optim — Module.

Optim.jl

Welcome to Optim.jl!

Optim.jl is a package used to solve continuous optimization problems. It is written in Julia for Julians to help take advantage of arbitrary number types, fast computation, and excellent automatic differentiation tools.

REPL help

? followed by an algorithm name (?BFGS), constructors (?Optim.Options) prints help to the terminal.

Documentation

Besides the help provided at the REPL, it is possible to find help and general documentation online at http://julianlsolvers.github.io/Optim.jl/stable/ .

Optim.OptimizationResults

Optim.OptimizationState

Optim.OptimizationTrace

Optim.Options

Optim.ParticleSwarm — Type.

Particle Swarm

Constructor

ParticleSwarm(; lower = [],
                upper = [],
                n_particles = 0)

The constructor takes 3 keywords:

lower = [], a vector of lower bounds, unbounded below if empty or Inf's

upper = [], a vector of upper bounds, unbounded above if empty or Inf's

n_particles = 0, the number of particles in the swarm, defaults to least three

Description

The Particle Swarm implementation in Optim.jl is the so-called Adaptive Particle Swarm algorithm in [1]. It attempts to improve global coverage and convergence by switching between four evolutionary states: exploration, exploitation, convergence, and jumping out. In the jumping out state it intentially tries to take the best particle and move it away from its (potentially and probably) local optimum, to improve the ability to find a global optimum. Of course, this comes a the cost of slower convergence, but hopefully converges to the global optimum as a result.

References

[1] Zhan, Zhang, and Chung. Adaptive particle swarm optimization, IEEE Transactions on Systems, Man, and Cybernetics, Part B: CyberneticsVolume 39, Issue 6 (2009): 1362-1381

Optim.ParticleSwarmState

Optim.PowerManifold — Type.

Multiple copies of the same manifold. Points are stored as innerdims x outerdims, e.g. the product of 2x2 Stiefel manifolds of dimension N x n would be a N x n x 2 x 2 matrix.

Optim.ProductManifold — Type.

Product of two manifolds {P = (x1,x2), x1 ∈ m1, x2 ∈ m2}. P is stored as a flat 1D array, and x1 is before x2 in memory. Use get_inner(m, x, {1,2}) to access x1 or x2 in their original format.

Optim.SAMIN — Type.

SAMIN

Constructor

SAMIN(; nt::Int = 5     # reduce temperature every nt*ns*dim(x_init) evaluations
        ns::Int = 5     # adjust bounds every ns*dim(x_init) evaluations
        rt::T = 0.9     # geometric temperature reduction factor: when temp changes, new temp is t=rt*t
        neps::Int = 5   # number of previous best values the final result is compared to
        f_tol::T = 1e-12 # the required tolerance level for function value comparisons
        x_tol::T = 1e-6 # the required tolerance level for x
        coverage_ok::Bool = false, # if false, increase temperature until initial parameter space is covered
        verbosity::Int = 1) # scalar: 0, 1, 2 or 3 (default = 1).

Description

The SAMIN method implements the Simulated Annealing algorithm for problems with bounds constrains a described in Goffe et. al. (1994) and Goffe (1996). The algorithm

References

Goffe, et. al. (1994) "Global Optimization of Statistical Functions with Simulated Annealing", Journal of Econometrics, V. 60, N. 1/2.

Goffe, William L. (1996) "SIMANN: A Global Optimization Algorithm using Simulated Annealing " Studies in Nonlinear Dynamics & Econometrics, Oct96, Vol. 1 Issue 3.

source

Optim.SecondOrderOptimizer

Optim.Simplexer

Optim.SimulatedAnnealing — Type.

SimulatedAnnealing

Constructor

SimulatedAnnealing(; neighbor = default_neighbor!,
                     temperature = log_temperature,
                     keep_best::Bool = true)

The constructor takes 3 keywords:

neighbor = a!(x_proposed, x_current), a mutating function of the current x,

and the proposed x

T = b(iteration), a function of the current iteration that returns a temperature

p = c(f_proposal, f_current, T), a function of the current temperature, current

function value and proposed function value that returns an acceptance probability

Description

Simulated Annealing is a derivative free method for optimization. It is based on the Metropolis-Hastings algorithm that was originally used to generate samples from a thermodynamics system, and is often used to generate draws from a posterior when doing Bayesian inference. As such, it is a probabilistic method for finding the minimum of a function, often over a quite large domains. For the historical reasons given above, the algorithm uses terms such as cooling, temperature, and acceptance probabilities.

Optim.SimulatedAnnealingState

Optim.Sphere — Type.

Spherical manifold {|x| = 1}.

Optim.Stiefel — Type.

N x n matrices with orthonormal columns, i.e. such that X'X = I. Special cases: N x 1 = sphere, N x N = orthogonal/unitary group. Stiefel() uses a SVD algorithm to compute the retraction. To use a Cholesky-based orthogonalization (faster but less stable), use Stiefel(:CholQR). When the function to be optimized depends only on the subspace X*X' spanned by a point X in the Stiefel manifold, first-order optimization algorithms are equivalent for the Stiefel and Grassmann manifold, so there is no separate Grassmann manifold.

Optim.Stiefel_CholQR

Optim.Stiefel_SVD

Optim.TwiceDifferentiable

Optim.TwiceDifferentiableConstraints

Optim.UnivariateOptimizationResults

Optim.UnivariateOptimizer

Optim.ZerothOrderOptimizer

Optim.ZerothOrderState

Optim._bv

Optim._lagrangian_linefunc

Optim._lagrangian_lineslope

Optim._updateA! — Function.

Update storage A[i,j] for NGMRES

Update storage A[i,j] for OACCEL

Optim._updateQ! — Function.

Update storage Q[i,j] and Q[j,i] for NGMRES

Update storage Q[i,j] and Q[j,i] for OACCEL

Optim._updateb! — Function.

Update storage b[i] for NGMRES

Update storage b[i] for OACCEL

Optim._updateη — Function.

Update value η for NGMRES

Update value η for OACCEL

Optim._updateξ! — Function.

Update storage ξ[i,:] for NGMRES

Update storage ξ[i,:] for OACCEL

Optim.abs_tol

Optim.add_default_opts!

Optim.after_while!

Optim.alphax

Optim.assess_convergence

Optim.backtrack_constrained

Optim.backtrack_constrained_grad

Optim.barrier_box

Optim.barrier_combined

Optim.barrier_grad! — Function.

barrier_grad!(bgrad, bounds, x, bstate, μ)
barrier_grad!(gsx, gsc, bounds, x, sx, sc, μ)

Compute the gradient of the barrier penalty at (x,sx,sc), where x is the current position, sx are the coordinate slack variables, and sc are the linear/nonlinear slack variables. bounds::ConstraintBounds holds the parsed bounds.

The result is added to gsx, and gsc, so these vectors need to be initialized appropriately.

Optim.barrier_method

Optim.barrier_value — Function.

barrier_value(constraints, state) -> val
barrier_value(bounds, x, sx, sc, μ) -> val

Compute the value of the barrier penalty at the current state, or at a position (x,sx,sc), where x is the current position, sx are the coordinate slack variables, and sc are the linear/nonlinear slack variables. bounds holds the parsed bounds.

Optim.bsv_seed

Optim.cE

Optim.cbar

Optim.centroid

Optim.centroid!

Optim.cg_steihaug!

Optim.check_hard_case_candidate

Optim.check_kwargs

Optim.common_trace!

Optim.compute_cost!

Optim.constant_temperature

Optim.converged

Optim.default_convergence_assessment

Optim.default_neighbor!

Optim.default_options

Optim.equality_grad! — Function.

equality_grad!(gx, gbstate, bounds, x, c, J, bstate)

Compute the gradient of equality_violation, storing the result in gx (an array) and gbstate::BarrierStateVars.

Optim.equality_grad_var!

Optim.equality_grad_λ!

Optim.equality_violation — Function.

equality_violation([f=identity], bounds, x, c, bstate) -> val
equality_violation([f=identity], bounds, x, c, sx, sc, λx, λc, λxE, λcE) -> val

Compute the sum of f(v_i), where v_i = λ_i*(target - observed) measures the difference between the current state and the equality-constrained state. bounds::ConstraintBounds holds the parsed bounds. x is the current position, sx are the coordinate slack variables, and sc are the linear/nonlinear slack variables. c holds the values of the linear-nonlinear constraints, and the λ arguments hold the Lagrange multipliers for x, sx, sc, and c respectively.

Optim.estimate_maxstep

Optim.eval

Optim.evalgrad

Optim.evalhess

Optim.f_abschange

Optim.f_calls

Optim.f_converged

Optim.f_increased

Optim.f_relchange

Optim.f_tol

Optim.f_trace

Optim.fallback_method

Optim.function_barrier

Optim.g_calls

Optim.g_converged

Optim.g_norm_trace

Optim.g_residual

Optim.g_tol

Optim.get_inner

Optim.get_mu_1

Optim.get_mu_2

Optim.get_mu_3

Optim.get_mu_4

Optim.get_swarm_state

Optim.gf

Optim.gradient_convergence_assessment

Optim.grid_search

Optim.h_calls

Optim.has_deprecated_fminbox

Optim.hessianI — Function.

hessianI(x, constraints, λcI, μ) -> h

Compute the hessian at x of the λcI-weighted sum of user-supplied constraint functions for just the inequalities. This also includes contributions from any variables with bounds at 0, since those do not cause introduction of a slack variable. Other (nonzero) box constraints do not contribute to h, because the hessian of x_i is zero. (They contribute indirectly via their slack variables.)

Optim.hessianI!

Optim.hessian_projections

Optim.housekeeping!

Optim.include

Optim.initial_convergence

Optim.initial_mu

Optim.initial_state

Optim.initialize_μ_λ! — Function.

initialize_μ_λ!(state, bounds, μ0=:auto, β=0.01)
initialize_μ_λ!(state, bounds, (Hobj,HcI), μ0=:auto, β=0.01)

Pick μ and λ to ensure that the equality constraints are satisfied locally (at the current state.x), and that the initial gradient including the barrier would be a descent direction for the problem without the barrier (μ = 0). This ensures that the search isn't pushed out of the basin of the user-supplied initial guess.

Upon entry, the objective function gradient, constraint values, and constraint jacobian must be set in state.g, state.c, and state.J respectively. If you also wish to ensure that the projection of Hessian is minimally-perturbed along the initial gradient, supply the hessian of the objective (Hobj) and

HcI = ∑_i (σ_i/s_i)∇∇ c_{Ii}

for the constraints. This can be obtained as

HcI = hessianI(state.x, constraints, 1./state.slack_c)

You can manually specify μ by supplying a numerical value for μ0. Whether calculated algorithmically or specified manually, the values of λ are set using the chosen μ.

Optim.is_smaller_eps

Optim.isfeasible — Function.

isfeasible(constraints, state) -> Bool
isfeasible(constraints, x, c) -> Bool
isfeasible(constraints, x) -> Bool
isfeasible(bounds, x, c) -> Bool

Return true if point x is feasible, given the constraints which specify bounds lx, ux, lc, and uc. x is feasible if

lx[i] <= x[i] <= ux[i]
lc[i] <= c[i] <= uc[i]

for all possible i.

Optim.isinterior — Function.

isinterior(constraints, state) -> Bool
isinterior(constraints, x, c) -> Bool
isinterior(constraints, x) -> Bool
isinterior(bounds, x, c) -> Bool

Return true if point x is on the interior of the allowed region, given the constraints which specify bounds lx, ux, lc, and uc. x is in the interior if

lx[i] < x[i] < ux[i]
lc[i] < c[i] < uc[i]

for all possible i.

Optim.iteration_limit_reached

Optim.iterations

Optim.jacobianE

Optim.jacobianI

Optim.jacobianx

Optim.lagrangian

Optim.lagrangian_fg!

Optim.lagrangian_fgvec!

Optim.lagrangian_linefunc

Optim.lagrangian_linefunc!

Optim.lagrangian_lineslope

Optim.lagrangian_lineslope!

Optim.lagrangian_vec

Optim.lambdaE

Optim.lambdaI

Optim.limit_X!

Optim.limits_box

Optim.linesearch_anon

Optim.log_temperature

Optim.loginf

Optim.lower_bound

Optim.ls_update!

Optim.maxdiff

Optim.maximize

Optim.maximizer

Optim.method

Optim.minimizer

Optim.minimum

Optim.mulhess

Optim.ngmres_oaccel_warned

Optim.nlprecon_post_accelerate!

Optim.nlprecon_post_optimize!

Optim.nmobjective

Optim.optimize

Optim.p_sq_norm

Optim.pack_vec

Optim.pack_vec!

Optim.parameters

Optim.perform_linesearch!

Optim.pick_best_f

Optim.pick_best_x

Optim.precondprep!

Optim.precondprepbox!

Optim.print_header

Optim.project_tangent

Optim.project_tangent!

Optim.promote_objtype

Optim.qrregularize!

Optim.rel_tol

Optim.res

Optim.reset_search_direction!

Optim.retract

Optim.retract!

Optim.setslack!

Optim.sigma

Optim.simplexer

Optim.slack

Optim.slopealpha

Optim.solve_slack!

Optim.solve_step!

Optim.solve_tr_subproblem!

Optim.trace

Optim.trace!

Optim.twoloop!

Optim.unpack_vec!

Optim.update!

Optim.update_fg!

Optim.update_g!

Optim.update_gtilde!

Optim.update_h!

Optim.update_state!

Optim.update_swarm!

Optim.update_swarm_params!

Optim.upper_bound

Optim.userλ — Function.

userλ(λcI, bounds) -> λ

Accumulates λcI into a vector λ ordered as the user-supplied constraint functions c. Upper and lower bounds are summed, weighted by σ. The resulting λ includes an overall negative sign so that this becomes the coefficient for the user-supplied hessian.

This is relevant only for the inequalities. If you want the λ for just the equalities, you can use λ[bounds.ceq] = λcE for a zero-filled λ.

Optim.x_abschange

Optim.x_converged

Optim.x_lower_trace

Optim.x_tol

Optim.x_trace

Optim.x_upper_trace