Docstrings

Unitful.@B

Unitful.@Np

Unitful.@_doctables

Unitful.@affineunit — Macro.

@affineunit(symb, abbr, offset)

Macro for easily defining affine units. offset gives the zero of the relative scale in terms of an absolute scale; the scaling is the same as the absolute scale. Example: @affineunit °C "°C" (27315//100)K is used internally to define degrees Celsius.

Unitful.@cNp

Unitful.@dB

Unitful.@derived_dimension — Macro.

@derived_dimension(name, dims)

Creates type aliases to allow dispatch on Unitful.Quantity, Unitful.Level, and Unitful.Units objects of a derived dimension, like area, which is just length squared. The type aliases are not exported.

dims is a Unitful.Dimensions object.

Returns nothing.

Usage examples:

@derived_dimension Area 𝐋^2 gives Area and AreaUnit type aliases

@derived_dimension Speed 𝐋/𝐓 gives Speed and SpeedUnit type aliases

Unitful.@dimension — Macro.

@dimension(symb, abbr, name)

Creates new dimensions. name will be used like an identifier in the type parameter for a Unitful.Dimension object. symb will be a symbol defined in the namespace from which this macro is called that is bound to a Unitful.Dimensions object. For most intents and purposes it is this object that the user would manipulate in doing dimensional analysis. The symbol is not exported.

This macro extends Unitful.abbr to display the new dimension in an abbreviated format using the string abbr.

Type aliases are created that allow the user to dispatch on Unitful.Quantity, Unitful.Level and Unitful.Units objects of the newly defined dimension. The type alias for quantities or levels is simply given by name, and the type alias for units is given by name*"Units", e.g. LengthUnits. Note that there is also LengthFreeUnits, for example, which is an alias for dispatching on FreeUnits with length dimensions. The aliases are not exported.

Finally, if you define new dimensions with @dimension you will need to specify a preferred unit for that dimension with Unitful.preferunits, otherwise promotion will not work with that dimension. This is done automatically in the @refunit macro.

Returns the Dimensions object to which symb is bound.

Usage example from src/pkgdefaults.jl: @dimension 𝐋 "𝐋" Length

Unitful.@logscale — Macro.

@logscale(symb,abbr,name,base,prefactor,irp)

Define a logarithmic scale. Unlike with units, there is no special treatment for power-of-ten prefixes (decibels and bels are defined separately). However, arbitrary bases are possible, and computationally appropriate log and exp functions are used in calculations when available (e.g. log2, log10 for base 2 and base 10, respectively).

This macro defines a MixedUnits object identified by symbol symb. This can be used to construct Gain types, e.g. 3*dB. Additionally, two other MixedUnits objects are defined, with symbols Symbol(symb,"_rp") and Symbol(symb,"_p"), e.g. dB_rp and dB_p, respectively. These objects serve nearly the same purpose, but have extra information in their type that indicates whether they should be considered as root-power ratios or power ratios upon conversion to pure numbers.

This macro also defines another macro available as @symb. For example, @dB in the case of decibels. This can be used to construct Level objects at parse time. Usage is like @dB 3V/1V.

prefactor is the prefactor out in front of the logarithm for this log scale. In all cases it is defined with respect to taking ratios of power quantities. Just divide by two if you want to refer to root-power / field quantities instead.

irp (short for "is root power?") specifies whether the logarithmic scale is defined with respect to ratios of power or root-power quantities. In short: use false if your scale is decibel-like, or true if your scale is neper-like.

Examples:

julia> using Unitful: V, W

julia> @logscale dΠ "dΠ" Decipies π 10 false
dΠ

julia> @dΠ π*V/1V
20.0 dΠ (1 V)

julia> dΠ(π*V, 1V)
20.0 dΠ (1 V)

julia> @dΠ π^2*V/1V
40.0 dΠ (1 V)

julia> @dΠ π*W/1W
10.0 dΠ (1 W)

Unitful.@logunit — Macro.

@logunit(symb, abbr, logscale, reflevel)

Defines a logarithmic unit. For examples see src/pkgdefaults.jl.

Unitful.@prefixed_unit_symbols — Macro.

@prefixed_unit_symbols(symb,name,dimension,basefactor)

Not called directly by the user. Given a unit symbol and a unit's name, will define units for each possible SI power-of-ten prefix on that unit.

Example: @prefixed_unit_symbols m Meter 𝐋 (1.0,1) results in nm, cm, m, km, ... all getting defined in the calling namespace.

Unitful.@refunit — Macro.

@refunit(symb, name, abbr, dimension, tf)

Define a reference unit, typically SI. Rather than define conversion factors between each and every unit of a given dimension, conversion factors are given between each unit and a reference unit, defined by this macro.

This macro extends Unitful.abbr so that the reference unit can be displayed in an abbreviated format. If tf == true, this macro generates symbols for every power of ten of the unit, using the standard SI prefixes. A dimension must be given (Unitful.Dimensions object) that specifies the dimension of the reference unit.

In principle, users can use this macro, but it probably does not make much sense to do so. If you define a new (probably unphysical) dimension using @dimension, then this macro will be necessary. With existing dimensions, you will almost certainly cause confusion if you use this macro. One potential use case would be to define a unit system without reference to SI. However, there's no explicit barrier to prevent attempting conversions between SI and this hypothetical unit system, which could yield unexpected results.

Note that this macro will also choose the new unit (no power-of-ten prefix) as the default unit for promotion given this dimension.

Returns the Unitful.FreeUnits object to which symb is bound.

Usage example: @refunit m "m" Meter 𝐋 true

This example, found in src/pkgdefaults.jl, generates km, m, cm, ...

Unitful.@u_str — Macro.

@u_str(unit)

String macro to easily recall units, dimensions, or quantities defined in unit modules that have been registered with Unitful.register.

If the same symbol is used for a Unitful.Units object defined in different modules, then the symbol found in the most recently registered module will be used.

Note that what goes inside must be parsable as a valid Julia expression. In other words, u"N m" will fail if you intended to write u"N*m".

Examples:

julia> 1.0u"m/s"
1.0 m s^-1

julia> 1.0u"N*m"
1.0 m N

julia> u"m,kg,s"
(m, kg, s)

julia> typeof(1.0u"m/s")
Quantity{Float64,𝐋*𝐓^-1,Unitful.FreeUnits{(m, s^-1),𝐋*𝐓^-1,nothing}}

julia> u"ħ"
1.0545718001391127e-34 J s

Unitful.@unit — Macro.

@unit(symb,abbr,name,equals,tf)

Define a unit. Rather than specifying a dimension like in @refunit, equals should be a Unitful.Quantity equal to one of the unit being defined. If tf == true, symbols will be made for each power-of-ten prefix.

Returns the Unitful.FreeUnits object to which symb is bound.

Usage example: @unit mi "mi" Mile (201168//125)*m false

This example will not generate kmi (kilomiles).

Unitful.@unit_symbols — Macro.

@unit_symbols(symb,name)

Not called directly by the user. Given a unit symbol and a unit's name, will define units without SI power-of-ten prefixes.

Example: @unit_symbols ft Foot 𝐋 results in ft getting defined but not kft.

Unitful.A

Unitful.AbsoluteScaleTemperature

Unitful.Acceleration

Unitful.AccelerationFreeUnits

Unitful.AccelerationUnits

Unitful.Action

Unitful.ActionFreeUnits

Unitful.ActionUnits

Unitful.Affine

Unitful.AffineError — Type.

struct AffineError <: Exception

An invalid operation was attempted with affine units / quantities.

Unitful.AffineQuantity — Type.

AffineQuantity{T,D,U} = Quantity{T,D,U} where U<:AffineUnits

Useful for dispatching on quantities that affine-transform under unit conversion, like absolute temperatures. Not exported.

Unitful.AffineUnits — Type.

AffineUnits{N,D,A} = Units{N,D,A} where A<:Affine

Useful for dispatching on unit objects that indicate a quantity should affine-transform under unit conversion, like absolute temperatures. Not exported.

Unitful.Amount

Unitful.AmountFreeUnits

Unitful.AmountUnits

Unitful.Area

Unitful.AreaFreeUnits

Unitful.AreaUnits

Unitful.B

Unitful.BField

Unitful.BFieldFreeUnits

Unitful.BFieldUnits

Unitful.B_p

Unitful.B_rp

Unitful.Ba

Unitful.Bel — Type.

struct LogInfo{N,B,P}

Describes a logarithmic unit. Type parameters include:

N: The name of the logarithmic unit, e.g. :Decibel, :Neper.

B: The base of the logarithm.

P: A prefactor to multiply the logarithm when the log is of a power ratio.

Unitful.Bq

Unitful.C

Unitful.Capacitance

Unitful.CapacitanceFreeUnits

Unitful.CapacitanceUnits

Unitful.Centineper — Type.

struct LogInfo{N,B,P}

Describes a logarithmic unit. Type parameters include:

N: The name of the logarithmic unit, e.g. :Decibel, :Neper.

B: The base of the logarithm.

P: A prefactor to multiply the logarithm when the log is of a power ratio.

Unitful.Charge

Unitful.ChargeFreeUnits

Unitful.ChargeUnits

Unitful.ContextUnits — Type.

struct ContextUnits{N,D,P,A} <: Units{N,D,A}

Instances of this object represent units, possibly combinations thereof. It is in most respects like FreeUnits{N,D,A}, except that the type parameter P is again a FreeUnits{M,D} type that specifies a preferred unit for promotion. See Advanced promotion mechanisms in the docs for details.

Unitful.Current

Unitful.CurrentFreeUnits

Unitful.CurrentUnits

Unitful.DField

Unitful.DFieldFreeUnits

Unitful.DFieldUnits

Unitful.Decibel — Type.

struct LogInfo{N,B,P}

Describes a logarithmic unit. Type parameters include:

N: The name of the logarithmic unit, e.g. :Decibel, :Neper.

B: The base of the logarithm.

P: A prefactor to multiply the logarithm when the log is of a power ratio.

Unitful.DefaultSymbols

Unitful.Density

Unitful.DensityFreeUnits

Unitful.DensityUnits

Unitful.Dimension — Type.

struct Dimension{D}
    power::Rational{Int}
end

Description of a dimension. The name of the dimension D is a symbol, e.g. :Length, :Time, :Mass, etc.

Dimension{D} objects are collected in a tuple, which is used for the type parameter N of a Dimensions{N} object.

Unitful.DimensionError — Type.

struct DimensionError <: Exception

Physical dimensions are inconsistent for the attempted operation.

Unitful.DimensionlessQuantity — Type.

DimensionlessQuantity{T,U} = Quantity{T, NoDims, U}

Useful for dispatching on Unitful.Quantity types that may have units but no dimensions. (Units with differing power-of-ten prefixes are not canceled out.)

Example:

julia> isa(1.0u"mV/V", DimensionlessQuantity)
true

Unitful.DimensionlessUnits — Type.

DimensionlessUnits{U}

Useful for dispatching on Unitful.Units types that have no dimensions.

Example: ```jldoctest julia> isa(Unitful.rad, DimensionlessUnits) true

Unitful.Dimensions — Type.

struct Dimensions{N} <: Unitlike

Instances of this object represent dimensions, possibly combinations thereof.

Unitful.DynamicViscosity

Unitful.DynamicViscosityFreeUnits

Unitful.DynamicViscosityUnits

Unitful.EA

Unitful.EBa

Unitful.EBq

Unitful.EC

Unitful.EF

Unitful.EField

Unitful.EFieldFreeUnits

Unitful.EFieldUnits

Unitful.EGal

Unitful.EGy

Unitful.EH

Unitful.EHz

Unitful.EHz2π

Unitful.EJ

Unitful.EK

Unitful.EL

Unitful.EN

Unitful.EP

Unitful.EPa

Unitful.ES

Unitful.ESt

Unitful.ESv

Unitful.ET

Unitful.ETorr

Unitful.EV

Unitful.EW

Unitful.EWb

Unitful.Eb

Unitful.Ebar

Unitful.Ecal

Unitful.Ecd

Unitful.Edyn

Unitful.EeV

Unitful.Eerg

Unitful.Eg

Unitful.Ekat

Unitful.El

Unitful.ElectricDipoleMoment

Unitful.ElectricDipoleMomentFreeUnits

Unitful.ElectricDipoleMomentUnits

Unitful.ElectricQuadrupoleMoment

Unitful.ElectricQuadrupoleMomentFreeUnits

Unitful.ElectricQuadrupoleMomentUnits

Unitful.ElectricalConductance

Unitful.ElectricalConductanceFreeUnits

Unitful.ElectricalConductanceUnits

Unitful.ElectricalConductivity

Unitful.ElectricalConductivityFreeUnits

Unitful.ElectricalConductivityUnits

Unitful.ElectricalResistance

Unitful.ElectricalResistanceFreeUnits

Unitful.ElectricalResistanceUnits

Unitful.ElectricalResistivity

Unitful.ElectricalResistivityFreeUnits

Unitful.ElectricalResistivityUnits

Unitful.Elm

Unitful.Elx

Unitful.Em

Unitful.Emol

Unitful.Energy

Unitful.EnergyFreeUnits

Unitful.EnergyUnits

Unitful.Erad

Unitful.Es

Unitful.Esr

Unitful.EΩ

Unitful.F

Unitful.FixedUnits — Type.

struct FixedUnits{N,D,A} <: Units{N,D,A} end

Instances of this object represent units, possibly combinations thereof. These are primarily intended for use when you would like to disable automatic unit conversions. See Advanced promotion mechanisms in the docs for details.

Unitful.Force

Unitful.ForceFreeUnits

Unitful.ForceUnits

Unitful.FreeOrContextUnits

Unitful.FreeUnits — Type.

struct FreeUnits{N,D,A} <: Units{N,D,A}

Instances of this object represent units, possibly combinations thereof. These behave like units have behaved in previous versions of Unitful, and provide a basic level of functionality that should be acceptable to most users. See Basic promotion mechanisms in the docs for details.

Example: the unit m is actually a singleton of type Unitful.FreeUnits{(Unitful.Unit{:Meter, 𝐋}(0, 1//1),), 𝐋, nothing}. After dividing by s, a singleton of type Unitful.FreeUnits{(Unitful.Unit{:Meter, 𝐋}(0, 1//1), Unitful.Unit{:Second, 𝐓}(0, -1//1)), 𝐋/𝐓, nothing} is returned.

Unitful.Frequency

Unitful.FrequencyFreeUnits

Unitful.FrequencyUnits

Unitful.G

Unitful.GA

Unitful.GBa

Unitful.GBq

Unitful.GC

Unitful.GF

Unitful.GGal

Unitful.GGy

Unitful.GH

Unitful.GHz

Unitful.GHz2π

Unitful.GJ

Unitful.GK

Unitful.GL

Unitful.GN

Unitful.GP

Unitful.GPa

Unitful.GS

Unitful.GSt

Unitful.GSv

Unitful.GT

Unitful.GTorr

Unitful.GV

Unitful.GW

Unitful.GWb

Unitful.Gain — Type.

struct Gain{L, S, T<:Real} <: LogScaled{L}

A logarithmic scale-based gain or attenuation factor. This type has one field, val::T. For example, given a gain of 20dB, we have val===20. This type differs from Unitful.Level in that val is stored after computing the logarithm.

Unitful.Gal

Unitful.Gb

Unitful.Gbar

Unitful.Gcal

Unitful.Gcd

Unitful.Gdyn

Unitful.GeV

Unitful.Gerg

Unitful.Gg

Unitful.Gkat

Unitful.Gl

Unitful.Glm

Unitful.Glx

Unitful.Gm

Unitful.Gmol

Unitful.Grad

Unitful.Gs

Unitful.Gsr

Unitful.Gy

Unitful.GΩ

Unitful.H

Unitful.HField

Unitful.HFieldFreeUnits

Unitful.HFieldUnits

Unitful.Hz

Unitful.Hz2π

Unitful.Inductance

Unitful.InductanceFreeUnits

Unitful.InductanceUnits

Unitful.InvalidOp

Unitful.IsRootPowerRatio

Unitful.J

Unitful.K

Unitful.KinematicViscosity

Unitful.KinematicViscosityFreeUnits

Unitful.KinematicViscosityUnits

Unitful.L

Unitful.Length

Unitful.LengthFreeUnits

Unitful.LengthUnits

Unitful.Level — Type.

struct Level{L, S, T<:Number} <: LogScaled{L}

A logarithmic scale-based level. Details about the logarithmic scale are encoded in L <: LogInfo. S is a reference quantity for the level, not a type. This type has one field, val::T, and the log of the ratio val/S is taken. This type differs from Unitful.Gain in that val is a linear quantity.

Unitful.LogInfo — Type.

struct LogInfo{N,B,P}

Describes a logarithmic unit. Type parameters include:

N: The name of the logarithmic unit, e.g. :Decibel, :Neper.

B: The base of the logarithm.

P: A prefactor to multiply the logarithm when the log is of a power ratio.

Unitful.LogScaled — Type.

abstract type LogScaled{L<:LogInfo} <: Number end

Abstract supertype of Unitful.Level and Unitful.Gain. It is only used in promotion to put levels and gains onto a common log scale.

Unitful.Luminosity

Unitful.LuminosityFreeUnits

Unitful.LuminosityUnits

Unitful.MA

Unitful.MBa

Unitful.MBq

Unitful.MC

Unitful.MF

Unitful.MGal

Unitful.MGy

Unitful.MH

Unitful.MHz

Unitful.MHz2π

Unitful.MJ

Unitful.MK

Unitful.ML

Unitful.MN

Unitful.MP

Unitful.MPa

Unitful.MS

Unitful.MSt

Unitful.MSv

Unitful.MT

Unitful.MTorr

Unitful.MV

Unitful.MW

Unitful.MWb

Unitful.MagneticDipoleMoment

Unitful.MagneticDipoleMomentFreeUnits

Unitful.MagneticDipoleMomentUnits

Unitful.MagneticFlux

Unitful.MagneticFluxFreeUnits

Unitful.MagneticFluxUnits

Unitful.Mass

Unitful.MassFreeUnits

Unitful.MassUnits

Unitful.Mb

Unitful.Mbar

Unitful.Mcal

Unitful.Mcd

Unitful.Mdyn

Unitful.MeV

Unitful.Merg

Unitful.Mg

Unitful.MixedUnits — Type.

struct MixedUnits{T<:LogScaled, U<:Units}

Struct for representing mixed logarithmic / linear units. Primarily useful as an intermediate for uconvert. T is <: Level or <: Gain.

Unitful.Mkat

Unitful.Ml

Unitful.Mlm

Unitful.Mlx

Unitful.Mm

Unitful.Mmol

Unitful.Momentum

Unitful.MomentumFreeUnits

Unitful.MomentumUnits

Unitful.Mrad

Unitful.Ms

Unitful.Msr

Unitful.MΩ

Unitful.N

Unitful.Na

Unitful.Neper — Type.

struct LogInfo{N,B,P}

Describes a logarithmic unit. Type parameters include:

N: The name of the logarithmic unit, e.g. :Decibel, :Neper.

B: The base of the logarithm.

P: A prefactor to multiply the logarithm when the log is of a power ratio.

Unitful.NoDims

Unitful.NoUnits

Unitful.Np

Unitful.Np_p

Unitful.Np_rp

Unitful.P

Unitful.PA

Unitful.PBa

Unitful.PBq

Unitful.PC

Unitful.PF

Unitful.PGal

Unitful.PGy

Unitful.PH

Unitful.PHz

Unitful.PHz2π

Unitful.PJ

Unitful.PK

Unitful.PL

Unitful.PN

Unitful.PP

Unitful.PPa

Unitful.PS

Unitful.PSt

Unitful.PSv

Unitful.PT

Unitful.PTorr

Unitful.PV

Unitful.PW

Unitful.PWb

Unitful.Pa

Unitful.Pb

Unitful.Pbar

Unitful.Pcal

Unitful.Pcd

Unitful.Pdyn

Unitful.PeV

Unitful.Perg

Unitful.Pg

Unitful.Pkat

Unitful.Pl

Unitful.Plm

Unitful.Plx

Unitful.Pm

Unitful.Pmol

Unitful.Power

Unitful.PowerFreeUnits

Unitful.PowerRatio

Unitful.PowerUnits

Unitful.Prad

Unitful.Pressure

Unitful.PressureFreeUnits

Unitful.PressureUnits

Unitful.Ps

Unitful.Psr

Unitful.PΩ

Unitful.Quantity — Type.

" struct Quantity{T,D,U} <: Number A quantity, which has dimensions and units specified in the type signature. The dimensions and units are allowed to be the empty set, in which case a dimensionless, unitless number results.

The type parameter T represents the numeric backing type. The type parameters D :: Unitful.Dimensions and U <: Unitful.Units. Of course, the dimensions follow from the units, but the type parameters are kept separate to permit convenient dispatch on dimensions.

Unitful.R

Unitful.Ra

Unitful.RelativeScaleTemperature

Unitful.RootPowerRatio

Unitful.R∞

Unitful.S

Unitful.ScalarQuantity — Type.

ScalarQuantity{T,D,U} = Quantity{T,D,U} where U<:ScalarUnits

Useful for dispatching on quantities that transform in the usual scalar way under unit conversion. Not exported.

Unitful.ScalarUnits — Type.

ScalarUnits{N,D} = Units{N,D,nothing}

Useful for dispatching on unit objects that indicate a quantity should transform in the usual scalar way under unit conversion. Not exported.

Unitful.St

Unitful.Sv

Unitful.T

Unitful.TA

Unitful.TBa

Unitful.TBq

Unitful.TC

Unitful.TF

Unitful.TGal

Unitful.TGy

Unitful.TH

Unitful.THz

Unitful.THz2π

Unitful.TJ

Unitful.TK

Unitful.TL

Unitful.TN

Unitful.TP

Unitful.TPa

Unitful.TS

Unitful.TSt

Unitful.TSv

Unitful.TT

Unitful.TTorr

Unitful.TV

Unitful.TW

Unitful.TWb

Unitful.Tb

Unitful.Tbar

Unitful.Tcal

Unitful.Tcd

Unitful.Tdyn

Unitful.TeV

Unitful.Temperature

Unitful.TemperatureFreeUnits

Unitful.TemperatureUnits

Unitful.Terg

Unitful.Tg

Unitful.Time

Unitful.TimeFreeUnits

Unitful.TimeUnits

Unitful.Tkat

Unitful.Tl

Unitful.Tlm

Unitful.Tlx

Unitful.Tm

Unitful.Tmol

Unitful.Torr

Unitful.Trad

Unitful.Ts

Unitful.Tsr

Unitful.TΩ

Unitful.Unit — Type.

struct Unit{U,D}
    tens::Int
    power::Rational{Int}
end

Description of a physical unit, including powers-of-ten prefixes and powers of the unit. The name of the unit is encoded in the type parameter U as a symbol, e.g. :Meter, :Second, :Gram, etc. The type parameter D is a Dimensions{N} object, for instance Unit{:Meter, 𝐋} or Unit{:Liter, 𝐋^3}. Note that the dimension information refers to the unit, not powers of the unit.

Unit{U,D} objects are almost never explicitly manipulated by the user. They are collected in a tuple, which is used for the type parameter N of a Units{N,D,A} object.

Unitful.Unitful

Unitful.Unitlike — Type.

abstract type Unitlike end

Represents units or dimensions. Dimensions are unit-like in the sense that they are not numbers but you can multiply or divide them and exponentiate by rationals.

Unitful.Units — Type.

abstract type Units{N,D,A} <: Unitlike end

Abstract supertype of all units objects, which can differ in their implementation details. A is a translation for affine quantities; for non-affine quantities it is nothing.

Unitful.V

Unitful.Velocity

Unitful.VelocityFreeUnits

Unitful.VelocityUnits

Unitful.Voltage

Unitful.VoltageFreeUnits

Unitful.VoltageUnits

Unitful.Volume

Unitful.VolumeFreeUnits

Unitful.VolumeUnits

Unitful.W

Unitful.Wavenumber

Unitful.WavenumberFreeUnits

Unitful.WavenumberUnits

Unitful.Wb

Unitful.YA

Unitful.YBa

Unitful.YBq

Unitful.YC

Unitful.YF

Unitful.YGal

Unitful.YGy

Unitful.YH

Unitful.YHz

Unitful.YHz2π

Unitful.YJ

Unitful.YK

Unitful.YL

Unitful.YN

Unitful.YP

Unitful.YPa

Unitful.YS

Unitful.YSt

Unitful.YSv

Unitful.YT

Unitful.YTorr

Unitful.YV

Unitful.YW

Unitful.YWb

Unitful.Yb

Unitful.Ybar

Unitful.Ycal

Unitful.Ycd

Unitful.Ydyn

Unitful.YeV

Unitful.Yerg

Unitful.Yg

Unitful.Ykat

Unitful.Yl

Unitful.Ylm

Unitful.Ylx

Unitful.Ym

Unitful.Ymol

Unitful.Yrad

Unitful.Ys

Unitful.Ysr

Unitful.YΩ

Unitful.Z0

Unitful.ZA

Unitful.ZBa

Unitful.ZBq

Unitful.ZC

Unitful.ZF

Unitful.ZGal

Unitful.ZGy

Unitful.ZH

Unitful.ZHz

Unitful.ZHz2π

Unitful.ZJ

Unitful.ZK

Unitful.ZL

Unitful.ZN

Unitful.ZP

Unitful.ZPa

Unitful.ZS

Unitful.ZSt

Unitful.ZSv

Unitful.ZT

Unitful.ZTorr

Unitful.ZV

Unitful.ZW

Unitful.ZWb

Unitful.Zb

Unitful.Zbar

Unitful.Zcal

Unitful.Zcd

Unitful.Zdyn

Unitful.ZeV

Unitful.Zerg

Unitful.Zg

Unitful.Zkat

Unitful.Zl

Unitful.Zlm

Unitful.Zlx

Unitful.Zm

Unitful.Zmol

Unitful.Zrad

Unitful.Zs

Unitful.Zsr

Unitful.ZΩ

Unitful._Quantity — Function.

Quantity(x::Number, y::Units)

Outer constructor for Quantitys. This is a generated function to avoid determining the dimensions of a given set of units each time a new quantity is made.

Unitful.__init__

Unitful._colon

Unitful._dimerr

Unitful._fma

Unitful._gconv

Unitful._isapprox

Unitful._isless

Unitful._lt

Unitful._muladd

Unitful._promote_unit

Unitful._range

Unitful.a

Unitful.aA

Unitful.aBa

Unitful.aBq

Unitful.aC

Unitful.aF

Unitful.aGal

Unitful.aGy

Unitful.aH

Unitful.aHz

Unitful.aHz2π

Unitful.aJ

Unitful.aK

Unitful.aL

Unitful.aN

Unitful.aP

Unitful.aPa

Unitful.aS

Unitful.aSt

Unitful.aSv

Unitful.aT

Unitful.aTorr

Unitful.aV

Unitful.aW

Unitful.aWb

Unitful.ab

Unitful.abar

Unitful.abbr — Function.

abbr(x) provides abbreviations for units or dimensions. Since a method should always be defined for each unit and dimension type, absence of a method for a specific unit or dimension type is likely an error. Consequently, we return ❓ for generic arguments to flag unexpected behavior.

Unitful.abs2_fast

Unitful.abs_fast

Unitful.absoluteunit — Function.

absoluteunit(::Units)
absoluteunit(::Quantity)

Given a unit or quantity, which may or may not be affine (e.g. °C), return the corresponding unit on the absolute temperature scale (e.g. K). Passing a Unitful.ContextUnits object will return another ContextUnits object with the same promotion unit, which may be an affine unit, so take care.

Unitful.ac

Unitful.acal

Unitful.acd

Unitful.adyn

Unitful.aeV

Unitful.aerg

Unitful.affinetranslation

Unitful.affineunit — Function.

affineunit(x::Quantity)

Returns a Unitful.Units object that can be used to construct affine quantities. Primarily, this is for relative temperatures (as opposed to absolute temperatures, which transform as usual under unit conversion). To use this function, pass the scale offset, e.g. affineunit(273.15K) yields a Celsius unit.

Unitful.ag

Unitful.akat

Unitful.al

Unitful.allowed_funcs

Unitful.alm

Unitful.alx

Unitful.am

Unitful.amol

Unitful.arad

Unitful.as

Unitful.asr

Unitful.atm

Unitful.aΩ

Unitful.b

Unitful.bar

Unitful.base

Unitful.basefactor — Function.

basefactor(x::Unit)

Specifies conversion factors to reference units. It returns a tuple. The first value is any irrational part of the conversion, and the second value is a rational component. This segregation permits exact conversions within unit systems that have no rational conversion to the reference units.

Unitful.basefactors

Unitful.btu

Unitful.c

Unitful.c0

Unitful.cA

Unitful.cBa

Unitful.cBq

Unitful.cC

Unitful.cF

Unitful.cGal

Unitful.cGy

Unitful.cH

Unitful.cHz

Unitful.cHz2π

Unitful.cJ

Unitful.cK

Unitful.cL

Unitful.cN

Unitful.cNp

Unitful.cNp_p

Unitful.cNp_rp

Unitful.cP

Unitful.cPa

Unitful.cS

Unitful.cSt

Unitful.cSv

Unitful.cT

Unitful.cTorr

Unitful.cV

Unitful.cW

Unitful.cWb

Unitful.cal

Unitful.cb

Unitful.cbar

Unitful.ccal

Unitful.ccd

Unitful.cd

Unitful.cdyn

Unitful.ceV

Unitful.cerg

Unitful.cg

Unitful.ckat

Unitful.cl

Unitful.clm

Unitful.clx

Unitful.cm

Unitful.cmol

Unitful.colon

Unitful.colonstartstop

Unitful.conj_fast

Unitful.convertr

Unitful.convertrp

Unitful.convfact — Function.

convfact(s::Units, t::Units)

Find the conversion factor from unit t to unit s, e.g. convfact(m,cm) = 0.01.

convfact{S}(s::Units{S}, t::Units{S})

Returns 1. (Avoids effort when unnecessary.)

Unitful.cos_fast

Unitful.crad

Unitful.cs

Unitful.csr

Unitful.cΩ

Unitful.d

Unitful.dA

Unitful.dB

Unitful.dBFS

Unitful.dBHz

Unitful.dBS

Unitful.dBSPL

Unitful.dBV

Unitful.dB_p

Unitful.dB_rp

Unitful.dBa

Unitful.dBm

Unitful.dBq

Unitful.dBu

Unitful.dBΩ

Unitful.dBμV

Unitful.dC

Unitful.dF

Unitful.dGal

Unitful.dGy

Unitful.dH

Unitful.dHz

Unitful.dHz2π

Unitful.dJ

Unitful.dK

Unitful.dL

Unitful.dN

Unitful.dP

Unitful.dPa

Unitful.dS

Unitful.dSt

Unitful.dSv

Unitful.dT

Unitful.dTorr

Unitful.dV

Unitful.dW

Unitful.dWb

Unitful.daA

Unitful.daBa

Unitful.daBq

Unitful.daC

Unitful.daF

Unitful.daGal

Unitful.daGy

Unitful.daH

Unitful.daHz

Unitful.daHz2π

Unitful.daJ

Unitful.daK

Unitful.daL

Unitful.daN

Unitful.daP

Unitful.daPa

Unitful.daS

Unitful.daSt

Unitful.daSv

Unitful.daT

Unitful.daTorr

Unitful.daV

Unitful.daW

Unitful.daWb

Unitful.dab

Unitful.dabar

Unitful.dacal

Unitful.dacd

Unitful.dadyn

Unitful.daeV

Unitful.daerg

Unitful.dag

Unitful.dakat

Unitful.dal

Unitful.dalm

Unitful.dalx

Unitful.dam

Unitful.damol

Unitful.darad

Unitful.das

Unitful.dasr

Unitful.daΩ

Unitful.db

Unitful.dbar

Unitful.dcal

Unitful.dcd

Unitful.ddyn

Unitful.deV

Unitful.derg

Unitful.dg

Unitful.dimension — Function.

dimension(x::Unit)

Returns a Unitful.Dimensions object describing the given unit x.

Unitful.dimtype

Unitful.dkat

Unitful.dl

Unitful.dlm

Unitful.dlx

Unitful.dm

Unitful.dmol

Unitful.dr

Unitful.drad

Unitful.ds

Unitful.dsr

Unitful.dyn

Unitful.dΩ

Unitful.eV

Unitful.erg

Unitful.eval

Unitful.expfn — Function.

expfn(x::LogInfo)

Returns the appropriate exponential function to use in calculations involving the logarithmic unit / quantity. For example, decibel-based units yield exp10, Neper-based yield exp, and so on. Returns x->(base)^x as a fallback.

Unitful.fA

Unitful.fBa

Unitful.fBq

Unitful.fC

Unitful.fF

Unitful.fGal

Unitful.fGy

Unitful.fH

Unitful.fHz

Unitful.fHz2π

Unitful.fJ

Unitful.fK

Unitful.fL

Unitful.fN

Unitful.fP

Unitful.fPa

Unitful.fS

Unitful.fSt

Unitful.fSv

Unitful.fT

Unitful.fTorr

Unitful.fV

Unitful.fW

Unitful.fWb

Unitful.fb

Unitful.fbar

Unitful.fcal

Unitful.fcd

Unitful.fdyn

Unitful.feV

Unitful.ferg

Unitful.fg

Unitful.fkat

Unitful.fl

Unitful.flm

Unitful.flx

Unitful.fm

Unitful.fmol

Unitful.frad

Unitful.fromlog

Unitful.fs

Unitful.fsr

Unitful.ft

Unitful.fΩ

Unitful.g

Unitful.gaintype

Unitful.ge

Unitful.genericunit — Function.

genericunit(::Units)

Given e.g. a FreeUnits{N,D,A}, ContextUnits{N,D,P,A}, or FixedUnits{N,D,A} object, return the type Units{N,D,A}.

Unitful.get_T

Unitful.gn

Unitful.gr

Unitful.h

Unitful.hA

Unitful.hBa

Unitful.hBq

Unitful.hC

Unitful.hF

Unitful.hGal

Unitful.hGy

Unitful.hH

Unitful.hHz

Unitful.hHz2π

Unitful.hJ

Unitful.hK

Unitful.hL

Unitful.hN

Unitful.hP

Unitful.hPa

Unitful.hS

Unitful.hSt

Unitful.hSv

Unitful.hT

Unitful.hTorr

Unitful.hV

Unitful.hW

Unitful.hWb

Unitful.ha

Unitful.hb

Unitful.hbar

Unitful.hcal

Unitful.hcd

Unitful.hdyn

Unitful.heV

Unitful.herg

Unitful.hg

Unitful.hkat

Unitful.hl

Unitful.hlm

Unitful.hlx

Unitful.hm

Unitful.hmol

Unitful.hr

Unitful.hrad

Unitful.hs

Unitful.hsr

Unitful.hΩ

Unitful.inch

Unitful.include

Unitful.inv_fast

Unitful.isrootpower

Unitful.isrootpower_dim

Unitful.isunitless

Unitful.k

Unitful.kA

Unitful.kBa

Unitful.kBq

Unitful.kC

Unitful.kF

Unitful.kGal

Unitful.kGy

Unitful.kH

Unitful.kHz

Unitful.kHz2π

Unitful.kJ

Unitful.kK

Unitful.kL

Unitful.kN

Unitful.kP

Unitful.kPa

Unitful.kS

Unitful.kSt

Unitful.kSv

Unitful.kT

Unitful.kTorr

Unitful.kV

Unitful.kW

Unitful.kWb

Unitful.kat

Unitful.kb

Unitful.kbar

Unitful.kcal

Unitful.kcd

Unitful.kdyn

Unitful.keV

Unitful.kerg

Unitful.kg

Unitful.kkat

Unitful.kl

Unitful.klm

Unitful.klx

Unitful.km

Unitful.kmol

Unitful.krad

Unitful.ks

Unitful.ksr

Unitful.kΩ

Unitful.l

Unitful.lb

Unitful.lbf

Unitful.leveltype

Unitful.linear — Function.

linear(x::Quantity)
linear(x::Level)
linear(x::Number) = x

Returns a quantity equivalent to x but without any logarithmic scales.

It is important to note that this operation will error for Quantity{<:Gain} types. This is for two reasons:

20dB could be interpreted as either a power or root-power ratio.

Even if -20dB/m were interpreted as, say, 0.01/m, this means something fundamentally different than -20dB/m. 0.01/m cannot be used to calculate exponential attenuation.

Unitful.lm

Unitful.logfn — Function.

logfn(x::LogInfo)

Returns the appropriate logarithm function to use in calculations involving the logarithmic unit / quantity. For example, decibel-based units yield log10, Neper-based yield ln, and so on. Returns x->log(base, x) as a fallback.

Unitful.logunit

Unitful.lx

Unitful.m

Unitful.mA

Unitful.mBa

Unitful.mBq

Unitful.mC

Unitful.mF

Unitful.mGal

Unitful.mGy

Unitful.mH

Unitful.mHz

Unitful.mHz2π

Unitful.mJ

Unitful.mK

Unitful.mL

Unitful.mN

Unitful.mP

Unitful.mPa

Unitful.mS

Unitful.mSt

Unitful.mSv

Unitful.mT

Unitful.mTorr

Unitful.mV

Unitful.mW

Unitful.mWb

Unitful.mb

Unitful.mbar

Unitful.mcal

Unitful.mcd

Unitful.mdyn

Unitful.me

Unitful.meV

Unitful.merg

Unitful.mg

Unitful.mi

Unitful.mil

Unitful.minute

Unitful.mkat

Unitful.ml

Unitful.mlm

Unitful.mlx

Unitful.mm

Unitful.mmol

Unitful.mn

Unitful.mol

Unitful.mp

Unitful.mrad

Unitful.ms

Unitful.msr

Unitful.mΩ

Unitful.nA

Unitful.nBa

Unitful.nBq

Unitful.nC

Unitful.nF

Unitful.nGal

Unitful.nGy

Unitful.nH

Unitful.nHz

Unitful.nHz2π

Unitful.nJ

Unitful.nK

Unitful.nL

Unitful.nN

Unitful.nP

Unitful.nPa

Unitful.nS

Unitful.nSt

Unitful.nSv

Unitful.nT

Unitful.nTorr

Unitful.nV

Unitful.nW

Unitful.nWb

Unitful.name

Unitful.nb

Unitful.nbar

Unitful.ncal

Unitful.ncd

Unitful.ndyn

Unitful.neV

Unitful.nerg

Unitful.ng

Unitful.nkat

Unitful.nl

Unitful.nlm

Unitful.nlx

Unitful.nm

Unitful.nmol

Unitful.nrad

Unitful.ns

Unitful.nsr

Unitful.numtype

Unitful.nΩ

Unitful.oz

Unitful.pA

Unitful.pBa

Unitful.pBq

Unitful.pC

Unitful.pF

Unitful.pGal

Unitful.pGy

Unitful.pH

Unitful.pHz

Unitful.pHz2π

Unitful.pJ

Unitful.pK

Unitful.pL

Unitful.pN

Unitful.pP

Unitful.pPa

Unitful.pS

Unitful.pSt

Unitful.pSv

Unitful.pT

Unitful.pTorr

Unitful.pV

Unitful.pW

Unitful.pWb

Unitful.pb

Unitful.pbar

Unitful.pcal

Unitful.pcd

Unitful.pdyn

Unitful.peV

Unitful.percent

Unitful.perg

Unitful.permille

Unitful.pg

Unitful.pkat

Unitful.pl

Unitful.plm

Unitful.plx

Unitful.pm

Unitful.pmol

Unitful.power

Unitful.prad

Unitful.prefactor

Unitful.preferunits — Function.

preferunits(u0::Units, u::Units...)

This function specifies the default fallback units for promotion. Units provided to this function must have a pure dimension of power 1, like 𝐋 or 𝐓 but not 𝐋/𝐓 or 𝐋^2. The function will complain if this is not the case. Additionally, the function will complain if you provide two units with the same dimension, as a courtesy to the user. Finally, you cannot use affine units such as °C with this function.

Once Unitful.upreferred has been called or quantities have been promoted, this function will appear to have no effect.

Usage example: preferunits(u"m,s,A,K,cd,kg,mol"...)

Unitful.prefix — Function.

prefix(x::Unit)

Returns a string representing the SI prefix for the power-of-ten held by this particular unit.

Unitful.prefixdict

Unitful.promote_to_derived — Function.

Unitful.promote_to_derived()

Defines promotion rules to use derived SI units in promotion for common dimensions of quantities:

J (joule) for energy

N (newton) for force

W (watt) for power

Pa (pascal) for pressure

C (coulomb) for charge

V (volt) for voltage

Ω (ohm) for resistance

F (farad) for capacitance

H (henry) for inductance

Wb (weber) for magnetic flux

T (tesla) for B-field

J*s (joule-second) for action

If you want this as default behavior (it was for versions of Unitful prior to 0.1.0), consider invoking this function in your .juliarc.jl file which is loaded when you open Julia. This function is not exported.

Unitful.promote_unit — Function.

promote_unit(::Units, ::Units...)

Given Units objects as arguments, this function returns a Units object appropriate for the result of promoting quantities which have these units. This function is kind of like promote_rule, except that it doesn't take types. It also does not return a tuple, but rather just a Unitful.Units object (or it throws an error).

Although we had used promote_rule for Units objects in prior versions of Unitful, this was always kind of a hack; it doesn't make sense to promote units directly for a variety of reasons.

Unitful.promotion

Unitful.ps

Unitful.psi

Unitful.psr

Unitful.pΩ

Unitful.q

Unitful.rad

Unitful.reflevel — Function.

reflevel(x::Level{L,S})
reflevel(::Type{Level{L,S}})
reflevel(::Type{Level{L,S,T}})

Returns the reference level, e.g.

julia> reflevel(3u"dBm")
1 mW

Unitful.register — Function.

register(unit_module::Module)

Makes the @u_str macro aware of units defined in new unit modules. By default, Unitful is itself a registered module. Note that Main is not, so if you define new units at the REPL, you will probably want to do Unitful.register(Main).

Example:

# somewhere in a custom units package...
module MyUnitsPackage
using Unitful

function __init__()
    ...
    Unitful.register(MyUnitsPackage)
end
end #module

Unitful.replace_value

Unitful.rpm

Unitful.rps

Unitful.s

Unitful.showrep — Function.

showrep(io::IO, x::Unit)

Show the unit, prefixing with any decimal prefix and appending the exponent as formatted by Unitful.superscript.

showrep(io::IO, x::Dimension)

Show the dimension, appending any exponent as formatted by Unitful.superscript.

Unitful.si_no_prefix

Unitful.si_prefixes

Unitful.sign_fast

Unitful.sin_fast

Unitful.slug

Unitful.sortexp — Function.

sortexp(xs)

Sort units to show positive exponents first.

Unitful.sr

Unitful.superscript — Function.

superscript(i::Rational)

Prints exponents.

Unitful.tan_fast

Unitful.tens

Unitful.tensfactor

Unitful.tolog

Unitful.u

Unitful.uconvert — Function.

uconvert(a::Units, x::Quantity{T,D,U}) where {T,D,U}

Convert a Unitful.Quantity to different units. The conversion will fail if the target units a have a different dimension than the dimension of the quantity x. You can use this method to switch between equivalent representations of the same unit, like N m and J.

Example:

julia> uconvert(u"hr",3602u"s")
1801//1800 hr
julia> uconvert(u"J",1.0u"N*m")
1.0 J

Unitful.uconvert_affine

Unitful.uconvertp — Function.

uconvertp(u::Units, x)
uconvertp(u::MixedUnits, x)

Generically, this is the same as Unitful.uconvert. In cases where unit conversion would be ambiguous without further information (e.g. uconvert(dB, 10)), uconvertp presumes ratios are of root-power quantities.

It is important to note that careless use of this function can lead to erroneous calculations. Consider Quantity{<:Gain} types: it is tempting to use this to transform -20dB/m into 0.1/m, however this means something fundamentally different than -20dB/m. Consider what happens when you try to compute exponential attenuation by multiplying 0.1/m by a length.

Examples:

julia> using Unitful

julia> uconvertp(u"dB", 10)
10.0 dB

julia> uconvertp(NoUnits, 20u"dB")
100.0

Unitful.uconvertrp — Function.

uconvertrp(u::Units, x)
uconvertrp(u::MixedUnits, x)

In most cases, this is the same as Unitful.uconvert. In cases where unit conversion would be ambiguous without further information (e.g. uconvert(dB, 10)), uconvertp presumes ratios are of power quantities.

It is important to note that careless use of this function can lead to erroneous calculations. Consider Quantity{<:Gain} types: it is tempting to use this to transform -20dB/m into 0.01/m, however this means something fundamentally different than -20dB/m. Consider what happens when you try to compute exponential attenuation by multiplying 0.01/m by a length.

Unitful.unit — Function.

unit(x::Quantity{T,D,U}) where {T,D,U}
unit(x::Type{Quantity{T,D,U}}) where {T,D,U}

Returns the units associated with a Quantity or Quantity type.

Examples:

julia> unit(1.0u"m") == u"m"
true

julia> unit(typeof(1.0u"m")) == u"m"
true

unit(x::Number)

Returns a Unitful.FreeUnits{(), NoDims} object to indicate that ordinary numbers have no units. This is a singleton, which we export as NoUnits. The unit is displayed as an empty string.

Examples:

julia> typeof(unit(1.0))
Unitful.FreeUnits{(),NoDims,nothing}

julia> typeof(unit(Float64))
Unitful.FreeUnits{(),NoDims,nothing}

julia> unit(1.0) == NoUnits
true

Unitful.unitmodules

Unitful.unwrap

Unitful.upreferred — Function.

upreferred(x::Dimensions)

Return units which are preferred for dimensions x. If you are using the factory defaults, this function will return a product of powers of base SI units (as Unitful.FreeUnits).

upreferred(x::Number)
upreferred(x::Quantity)

Unit-convert x to units which are preferred for the dimensions of x. If you are using the factory defaults, this function will unit-convert to a product of powers of base SI units. If quantity x has Unitful.ContextUnits(y,z), the resulting quantity will have units ContextUnits(z,z).

upreferred(x::Units)

Return units which are preferred for the dimensions of x, which may or may not be equal to x, as specified by the preferunits function. If you are using the factory defaults, this function will return a product of powers of base SI units.

Unitful.ustrcheck_bool

Unitful.ustrip — Function.

ustrip(x::Number)
ustrip(x::Quantity)

Returns the number out in front of any units. The value of x may differ from the number out front of the units in the case of dimensionless quantities, e.g. 1m/mm != 1. See uconvert and the example below. Because the units are removed, information may be lost and this should be used with some care.

This function is mainly intended for compatibility with packages that don't know how to handle quantities.

julia> ustrip(2u"μm/m") == 2
true

julia> uconvert(NoUnits, 2u"μm/m") == 2//1000000
true

ustrip(x::Array{Q}) where {Q <: Quantity}

Strip units from an Array by reinterpreting to type T. The resulting Array is a not a copy, but rather a unit-stripped view into array x. Because the units are removed, information may be lost and this should be used with some care.

This function is provided primarily for compatibility purposes; you could pass the result to PyPlot, for example.

julia> a = [1u"m", 2u"m"]
2-element Array{Quantity{Int64,𝐋,Unitful.FreeUnits{(m,),𝐋,nothing}},1}:
 1 m
 2 m

julia> b = ustrip(a)
2-element reinterpret(Int64, ::Array{Quantity{Int64,𝐋,Unitful.FreeUnits{(m,),𝐋,nothing}},1}):
 1
 2

julia> a[1] = 3u"m"; b
2-element reinterpret(Int64, ::Array{Quantity{Int64,𝐋,Unitful.FreeUnits{(m,),𝐋,nothing}},1}):
 3
 2

ustrip(A::Diagonal)
ustrip(A::Bidiagonal)
ustrip(A::Tridiagonal)
ustrip(A::SymTridiagonal)

Strip units from various kinds of matrices by calling ustrip on the underlying vectors.

Unitful.wk

Unitful.yA

Unitful.yBa

Unitful.yBq

Unitful.yC

Unitful.yF

Unitful.yGal

Unitful.yGy

Unitful.yH

Unitful.yHz

Unitful.yHz2π

Unitful.yJ

Unitful.yK

Unitful.yL

Unitful.yN

Unitful.yP

Unitful.yPa

Unitful.yS

Unitful.ySt

Unitful.ySv

Unitful.yT

Unitful.yTorr

Unitful.yV

Unitful.yW

Unitful.yWb

Unitful.yb

Unitful.ybar

Unitful.ycal

Unitful.ycd

Unitful.yd

Unitful.ydyn

Unitful.yeV

Unitful.yerg

Unitful.yg

Unitful.ykat

Unitful.yl

Unitful.ylm

Unitful.ylx

Unitful.ym

Unitful.ymol

Unitful.yrad

Unitful.ys

Unitful.ysr

Unitful.yΩ

Unitful.zA

Unitful.zBa

Unitful.zBq

Unitful.zC

Unitful.zF

Unitful.zGal

Unitful.zGy

Unitful.zH

Unitful.zHz

Unitful.zHz2π

Unitful.zJ

Unitful.zK

Unitful.zL

Unitful.zN

Unitful.zP

Unitful.zPa

Unitful.zS

Unitful.zSt

Unitful.zSv

Unitful.zT

Unitful.zTorr

Unitful.zV

Unitful.zW

Unitful.zWb

Unitful.zb

Unitful.zbar

Unitful.zcal

Unitful.zcd

Unitful.zdyn

Unitful.zeV

Unitful.zerg

Unitful.zg

Unitful.zkat

Unitful.zl

Unitful.zlm

Unitful.zlx

Unitful.zm

Unitful.zmol

Unitful.zrad

Unitful.zs

Unitful.zsr

Unitful.zΩ

Unitful.°

Unitful.°C

Unitful.°F

Unitful.ħ

Unitful.Φ0

Unitful.Ω

Unitful.ε0

Unitful.μ0

Unitful.μA

Unitful.μB

Unitful.μBa

Unitful.μBq

Unitful.μC

Unitful.μF

Unitful.μGal

Unitful.μGy

Unitful.μH

Unitful.μHz

Unitful.μHz2π

Unitful.μJ

Unitful.μK

Unitful.μL

Unitful.μN

Unitful.μP

Unitful.μPa

Unitful.μS

Unitful.μSt

Unitful.μSv

Unitful.μT

Unitful.μTorr

Unitful.μV

Unitful.μW

Unitful.μWb

Unitful.μb

Unitful.μbar

Unitful.μcal

Unitful.μcd

Unitful.μdyn

Unitful.μeV

Unitful.μerg

Unitful.μg

Unitful.μkat

Unitful.μl

Unitful.μlm

Unitful.μlx

Unitful.μm

Unitful.μmol

Unitful.μrad

Unitful.μs

Unitful.μsr

Unitful.μΩ

Unitful.σ

Unitful.ϵ0

Unitful.𝐈

Unitful.𝐉

Unitful.𝐋

Unitful.𝐌

Unitful.𝐍

Unitful.𝐓

Unitful.𝚯