Docstrings

PlotlyJS.@L_mstr

PlotlyJS.@L_str

PlotlyJS.AbstractLayout

PlotlyJS.AbstractTrace

PlotlyJS.Cycler

PlotlyJS.GenericTrace

PlotlyJS.Layout

PlotlyJS.Plot

PlotlyJS.PlotlyBase

PlotlyJS.PlotlyJS

PlotlyJS.PlotlyJSDisplay

PlotlyJS.STYLES

PlotlyJS.Shape

PlotlyJS.Style

PlotlyJS.SyncPlot

PlotlyJS.__init__

PlotlyJS._js_cdn_path

PlotlyJS._js_path

PlotlyJS._mathjax_cdn_path

PlotlyJS._pkg_root

PlotlyJS.add_recession_bands!

PlotlyJS.addtraces

PlotlyJS.addtraces!

PlotlyJS.area

PlotlyJS.attr

PlotlyJS.bar

PlotlyJS.box

PlotlyJS.candlestick

PlotlyJS.carpet

PlotlyJS.choropleth

PlotlyJS.circle

PlotlyJS.cone

PlotlyJS.contour

PlotlyJS.contourcarpet

PlotlyJS.deletetraces

PlotlyJS.deletetraces!

PlotlyJS.display_blink

PlotlyJS.docs

PlotlyJS.download_image

PlotlyJS.eval

PlotlyJS.extendtraces

PlotlyJS.extendtraces!

PlotlyJS.fork

PlotlyJS.heatmap

PlotlyJS.heatmapgl

PlotlyJS.histogram

PlotlyJS.histogram2d

PlotlyJS.histogram2dcontour

PlotlyJS.hline

PlotlyJS.html_body

PlotlyJS.include

PlotlyJS.js_default

PlotlyJS.json

PlotlyJS.line

PlotlyJS.mesh3d

PlotlyJS.movetraces

PlotlyJS.movetraces!

PlotlyJS.ohlc

PlotlyJS.parcoords

PlotlyJS.path

PlotlyJS.pie

PlotlyJS.plot

PlotlyJS.pointcloud

PlotlyJS.prependtraces

PlotlyJS.prependtraces!

PlotlyJS.purge

PlotlyJS.purge!

PlotlyJS.react

PlotlyJS.react!

PlotlyJS.rect

PlotlyJS.redraw

PlotlyJS.redraw!

PlotlyJS.relayout

PlotlyJS.relayout!

PlotlyJS.restyle

PlotlyJS.restyle!

PlotlyJS.sankey

PlotlyJS.savefig

PlotlyJS.savehtml — Function.

savehtml(io::IO, p::Union{Plot,SyncPlot}, js::Symbol=js_default[])
savehtml(p::Union{Plot,SyncPlot}, fn::AbstractString, js::Symbol=js_default[])

Save plot to standalone html file suitable for including in a website or opening in a browser

Can either be written to an arbitrary IO stream, or saved to a file noted with a string fn.

The js argument can be one of

:local: Reference the local plotly.js file included in this Julia package Pros: small file size, offline viewing. Cons: Can't share with others or move to different machine..

:remote: Reference plotly.js from a CDN. Pros small file size, move to other machine. Cons: need internet access to fetch from CDN

:embed: Embed the entirety of your local copy of plotly.js in the outputted file. Pros: offline viewing, move to other machine. Con: large file size (adds about 2.7 MB)

The default is :local

PlotlyJS.savejson

PlotlyJS.scatter

PlotlyJS.scatter3d

PlotlyJS.scattercarpet

PlotlyJS.scattergeo

PlotlyJS.scattergl

PlotlyJS.scattermapbox

PlotlyJS.scatterpolar

PlotlyJS.scatterpolargl

PlotlyJS.scatterternary

PlotlyJS.send_command

PlotlyJS.splom

PlotlyJS.stem

PlotlyJS.streamtube

PlotlyJS.style

PlotlyJS.surface

PlotlyJS.table

PlotlyJS.to_image

PlotlyJS.update

PlotlyJS.update!

PlotlyJS.use_style!

PlotlyJS.violin

PlotlyJS.vline