Docstrings

HDF5.@checked_lib

HDF5.@read

HDF5.@write

HDF5.ASCIIChar

HDF5.ASCII_ATTRIBUTE_PROPERTIES

HDF5.ASCII_LINK_PROPERTIES

HDF5.C_time_t

HDF5.CharType

HDF5.DEFAULT_PROPERTIES

HDF5.DataFile

HDF5.DatasetOrAttribute

HDF5.EmptyArray

HDF5.FILTER_BLOSC

HDF5.FILTER_BLOSC_VERSION

HDF5.FixedArray

HDF5.H5D_CHUNKED

HDF5.H5D_COMPACT

HDF5.H5D_CONTIGUOUS

HDF5.H5E_DEFAULT

HDF5.H5FD_MPIO_CHUNK_DEFAULT

HDF5.H5FD_MPIO_CHUNK_MULTI_IO

HDF5.H5FD_MPIO_CHUNK_ONE_IO

HDF5.H5FD_MPIO_COLLECTIVE

HDF5.H5FD_MPIO_COLLECTIVE_IO

HDF5.H5FD_MPIO_INDEPENDENT

HDF5.H5FD_MPIO_INDIVIDUAL_IO

HDF5.H5F_ACC_CREAT

HDF5.H5F_ACC_DEBUG

HDF5.H5F_ACC_EXCL

HDF5.H5F_ACC_RDONLY

HDF5.H5F_ACC_RDWR

HDF5.H5F_ACC_SWMR_READ

HDF5.H5F_ACC_SWMR_WRITE

HDF5.H5F_ACC_TRUNC

HDF5.H5F_CLOSE_DEFAULT

HDF5.H5F_CLOSE_SEMI

HDF5.H5F_CLOSE_STRONG

HDF5.H5F_CLOSE_WEAK

HDF5.H5F_LIBVER_EARLIEST

HDF5.H5F_LIBVER_LATEST

HDF5.H5F_LIBVER_V110

HDF5.H5F_LIBVER_V18

HDF5.H5F_OBJ_ALL

HDF5.H5F_OBJ_ATTR

HDF5.H5F_OBJ_DATASET

HDF5.H5F_OBJ_DATATYPE

HDF5.H5F_OBJ_FILE

HDF5.H5F_OBJ_GROUP

HDF5.H5F_OBJ_LOCAL

HDF5.H5F_SCOPE_GLOBAL

HDF5.H5F_SCOPE_LOCAL

HDF5.H5GINFO_TEMP_ARRAY

HDF5.H5Ginfo

HDF5.H5I_ATTR

HDF5.H5I_DATASET

HDF5.H5I_DATASPACE

HDF5.H5I_DATATYPE

HDF5.H5I_FILE

HDF5.H5I_GROUP

HDF5.H5I_REFERENCE

HDF5.H5LInfo

HDF5.H5L_TYPE_EXTERNAL

HDF5.H5L_TYPE_HARD

HDF5.H5L_TYPE_SOFT

HDF5.H5OINFO_TEMP_ARRAY

HDF5.H5O_TYPE_DATASET

HDF5.H5O_TYPE_GROUP

HDF5.H5O_TYPE_NAMED_DATATYPE

HDF5.H5Oinfo

HDF5.H5P_ATTRIBUTE_CREATE

HDF5.H5P_DATASET_ACCESS

HDF5.H5P_DATASET_CREATE

HDF5.H5P_DATASET_XFER

HDF5.H5P_DATATYPE_ACCESS

HDF5.H5P_DATATYPE_CREATE

HDF5.H5P_DEFAULT

HDF5.H5P_FILE_ACCESS

HDF5.H5P_FILE_CREATE

HDF5.H5P_FILE_MOUNT

HDF5.H5P_GROUP_ACCESS

HDF5.H5P_GROUP_CREATE

HDF5.H5P_LINK_ACCESS

HDF5.H5P_LINK_CREATE

HDF5.H5P_OBJECT_COPY

HDF5.H5P_OBJECT_CREATE

HDF5.H5P_STRING_CREATE

HDF5.H5RDEREFERENCE

HDF5.H5R_DATASET_REGION

HDF5.H5R_DSET_REG_REF_BUF_SIZE

HDF5.H5R_OBJECT

HDF5.H5R_OBJ_REF_BUF_SIZE

HDF5.H5S_ALL

HDF5.H5S_NULL

HDF5.H5S_SCALAR

HDF5.H5S_SELECT_AND

HDF5.H5S_SELECT_APPEND

HDF5.H5S_SELECT_NOTA

HDF5.H5S_SELECT_NOTB

HDF5.H5S_SELECT_OR

HDF5.H5S_SELECT_PREPEND

HDF5.H5S_SELECT_SET

HDF5.H5S_SELECT_XOR

HDF5.H5S_SIMPLE

HDF5.H5S_UNLIMITED

HDF5.H5T_ARRAY

HDF5.H5T_BITFIELD

HDF5.H5T_COMPOUND

HDF5.H5T_CSET_ASCII

HDF5.H5T_CSET_UTF8

HDF5.H5T_C_S1

HDF5.H5T_DIR_ASCEND

HDF5.H5T_DIR_DESCEND

HDF5.H5T_ENUM

HDF5.H5T_FLOAT

HDF5.H5T_IEEE_F32BE

HDF5.H5T_IEEE_F32LE

HDF5.H5T_IEEE_F64BE

HDF5.H5T_IEEE_F64LE

HDF5.H5T_INTEGER

HDF5.H5T_NATIVE_B8

HDF5.H5T_NATIVE_DOUBLE

HDF5.H5T_NATIVE_FLOAT

HDF5.H5T_NATIVE_INT16

HDF5.H5T_NATIVE_INT32

HDF5.H5T_NATIVE_INT64

HDF5.H5T_NATIVE_INT8

HDF5.H5T_NATIVE_UINT16

HDF5.H5T_NATIVE_UINT32

HDF5.H5T_NATIVE_UINT64

HDF5.H5T_NATIVE_UINT8

HDF5.H5T_OPAQUE

HDF5.H5T_REFERENCE

HDF5.H5T_SGN_2

HDF5.H5T_SGN_NONE

HDF5.H5T_STD_I16BE

HDF5.H5T_STD_I16LE

HDF5.H5T_STD_I32BE

HDF5.H5T_STD_I32LE

HDF5.H5T_STD_I64BE

HDF5.H5T_STD_I64LE

HDF5.H5T_STD_I8BE

HDF5.H5T_STD_I8LE

HDF5.H5T_STD_REF_DSETREG

HDF5.H5T_STD_REF_OBJ

HDF5.H5T_STD_U16BE

HDF5.H5T_STD_U16LE

HDF5.H5T_STD_U32BE

HDF5.H5T_STD_U32LE

HDF5.H5T_STD_U64BE

HDF5.H5T_STD_U64LE

HDF5.H5T_STD_U8BE

HDF5.H5T_STD_U8LE

HDF5.H5T_STRING

HDF5.H5T_STR_NULLPAD

HDF5.H5T_STR_NULLTERM

HDF5.H5T_STR_SPACEPAD

HDF5.H5T_TIME

HDF5.H5T_VARIABLE

HDF5.H5T_VLEN

HDF5.H5T_class_t

HDF5.H5Z_CLASS_T_VERS

HDF5.H5Z_FLAG_OPTIONAL

HDF5.H5Z_FLAG_REVERSE

HDF5.H5Z_class2_t

HDF5.H5Z_filter_t

HDF5.H5_INDEX_CRT_ORDER

HDF5.H5_INDEX_NAME

HDF5.H5_INDEX_UNKNOWN

HDF5.H5_ITER_DEC

HDF5.H5_ITER_INC

HDF5.H5_ITER_N

HDF5.H5_ITER_NATIVE

HDF5.H5_ITER_UNKNOWN

HDF5.HDF5

HDF5.HDF5Attribute

HDF5.HDF5Attributes

HDF5.HDF5BitsKind

HDF5.HDF5ChunkStorage

HDF5.HDF5Compound

HDF5.HDF5Dataset

HDF5.HDF5Dataspace

HDF5.HDF5Datatype

HDF5.HDF5File

HDF5.HDF5Group

HDF5.HDF5Object

HDF5.HDF5Opaque

HDF5.HDF5Properties

HDF5.HDF5ReferenceObj

HDF5.HDF5ReferenceObj_NULL

HDF5.HDF5Scalar

HDF5.HDF5Vlen

HDF5.HVL_SIZE

HDF5.Haddr

HDF5.Herr

HDF5.Hid

HDF5.Hmetainfo

HDF5.Hmpih

HDF5.Hmpih32

HDF5.Hmpih64

HDF5.Hsize

HDF5.Hssize

HDF5.Htri

HDF5.Hvl_t

HDF5.LENGTH_TEMP_ARRAY

HDF5.REF_TEMP_ARRAY

HDF5.ScalarOrString

HDF5.UTF8Char

HDF5.UTF8_ATTRIBUTE_PROPERTIES

HDF5.UTF8_LINK_PROPERTIES

HDF5.__init__

HDF5._a_create

HDF5._a_write

HDF5._attr_properties

HDF5._d_create

HDF5._d_write

HDF5._dataspace

HDF5._getindex

HDF5._link_properties

HDF5._set_blosc_values

HDF5._setindex!

HDF5.a_create

HDF5.a_delete

HDF5.a_open

HDF5.a_read

HDF5.a_write

HDF5.attrs

HDF5.atype

HDF5.blosc_chunkdims

HDF5.blosc_filter

HDF5.blosc_flags_

HDF5.blosc_name

HDF5.blosc_nelements_

HDF5.blosc_set_local

HDF5.blosc_values

HDF5.ccallexpr

HDF5.ccallsyms

HDF5.chartype

HDF5.checkprops

HDF5.checkvalid

HDF5.chunked_props

HDF5.create_external — Function.

create_external(source::Union{HDF5File, HDF5Group}, source_relpath, target_filename, target_path;
                lcpl_id=HDF5.H5P_DEFAULT, lapl_id=HDF5.H5P.DEFAULT)

Create an external link such that source[source_relpath] points to target_path within the file with path target_filename; Calls [H5Lcreate_external](https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-CreateExternal).

HDF5.cset

HDF5.d_create

HDF5.d_create_external

HDF5.d_open

HDF5.d_read

HDF5.d_write

HDF5.dataspace

HDF5.datatype

HDF5.depsfile

HDF5.do_write_chunk

HDF5.eval

HDF5.exists

HDF5.fd

HDF5.file

HDF5.filename

HDF5.funcdecexpr

HDF5.g_create

HDF5.g_open

HDF5.get_chunk

HDF5.get_create_properties

HDF5.get_datasets — Function.

get_datasets(file::HDF5File) -> datasets::Vector{HDF5Dataset}

Get all the datasets in an hdf5 file without loading the data.

HDF5.get_datasets!

HDF5.get_dims — Function.

get_dims(dset::HDF5Dataset)

Get the array dimensions from a dataset and return a tuple of dims and maxdims.

HDF5.get_fclose_degree

HDF5.get_libver_bounds

HDF5.get_userblock

HDF5.gettype

HDF5.h5_close

HDF5.h5_dont_atexit

HDF5.h5_garbage_collect

HDF5.h5_get_libversion

HDF5.h5_mpihandle

HDF5.h5_open

HDF5.h5_set_free_list_limits

HDF5.h5a_close

HDF5.h5a_create

HDF5.h5a_create_by_name

HDF5.h5a_delete

HDF5.h5a_delete_by_idx

HDF5.h5a_delete_by_name

HDF5.h5a_exists

HDF5.h5a_exists_by_name

HDF5.h5a_get_create_plist

HDF5.h5a_get_name

HDF5.h5a_get_name_by_idx

HDF5.h5a_get_space

HDF5.h5a_get_type

HDF5.h5a_open

HDF5.h5a_read

HDF5.h5a_write

HDF5.h5d_close

HDF5.h5d_create

HDF5.h5d_flush

HDF5.h5d_get_access_plist

HDF5.h5d_get_create_plist

HDF5.h5d_get_offset

HDF5.h5d_get_space

HDF5.h5d_get_type

HDF5.h5d_oappend

HDF5.h5d_open

HDF5.h5d_read

HDF5.h5d_refresh

HDF5.h5d_set_extent

HDF5.h5d_vlen_get_buf_size

HDF5.h5d_vlen_reclaim

HDF5.h5d_write

HDF5.h5do_write_chunk

HDF5.h5e_set_auto

HDF5.h5f_close

HDF5.h5f_create

HDF5.h5f_flush

HDF5.h5f_get_access_plist

HDF5.h5f_get_create_plist

HDF5.h5f_get_name

HDF5.h5f_get_obj_ids

HDF5.h5f_get_vfd_handle

HDF5.h5f_is_hdf5

HDF5.h5f_open

HDF5.h5g_close

HDF5.h5g_create

HDF5.h5g_get_create_plist

HDF5.h5g_get_info

HDF5.h5g_get_num_objs

HDF5.h5g_get_objname_by_idx

HDF5.h5g_open

HDF5.h5i_dec_ref

HDF5.h5i_get_file_id

HDF5.h5i_get_name

HDF5.h5i_get_ref

HDF5.h5i_get_type

HDF5.h5i_is_valid

HDF5.h5l_create_external

HDF5.h5l_create_hard

HDF5.h5l_create_soft

HDF5.h5l_delete

HDF5.h5l_exists

HDF5.h5l_get_info

HDF5.h5o_close

HDF5.h5o_copy

HDF5.h5o_get_info

HDF5.h5o_open

HDF5.h5o_open_by_addr

HDF5.h5o_open_by_idx

HDF5.h5object

HDF5.h5open — Function.

h5open(filename::AbstractString, mode::AbstractString="r"; swmr=false)

Open or create an HDF5 file where mode is one of:

"r" read only

"r+" read and write

"cw" read and write, create file if not existing, do not truncate

"w" read and write, create a new file (destroys any existing contents)

Pass swmr=true to enable (Single Writer Multiple Reader) SWMR write access for "w" and "r+", or SWMR read access for "r".

function h5open(f::Function, args...; swmr=false)

Apply the function f to the result of h5open(args...;kwargs...) and close the resulting HDF5File upon completion. For example with a do block:

h5open("foo.h5","w") do h5
    h5["foo"]=[1,2,3]
end

HDF5.h5p_close

HDF5.h5p_create

HDF5.h5p_get_chunk

HDF5.h5p_get_driver

HDF5.h5p_get_driver_info

HDF5.h5p_get_dxpl_mpio

HDF5.h5p_get_fapl_mpio

HDF5.h5p_get_fapl_mpio32

HDF5.h5p_get_fapl_mpio64

HDF5.h5p_get_fclose_degree

HDF5.h5p_get_layout

HDF5.h5p_get_userblock

HDF5.h5p_set_blosc

HDF5.h5p_set_char_encoding

HDF5.h5p_set_chunk

HDF5.h5p_set_create_intermediate_group

HDF5.h5p_set_deflate

HDF5.h5p_set_dxpl_mpio

HDF5.h5p_set_external

HDF5.h5p_set_fapl_mpio

HDF5.h5p_set_fapl_mpio32

HDF5.h5p_set_fapl_mpio64

HDF5.h5p_set_fclose_degree

HDF5.h5p_set_layout

HDF5.h5p_set_libver_bounds

HDF5.h5p_set_local_heap_size_hint

HDF5.h5p_set_obj_track_times

HDF5.h5p_set_shuffle

HDF5.h5p_set_userblock

HDF5.h5r_create

HDF5.h5r_dereference

HDF5.h5r_get_obj_type

HDF5.h5r_get_region

HDF5.h5read

HDF5.h5readattr

HDF5.h5rewrite

HDF5.h5s_close

HDF5.h5s_copy

HDF5.h5s_create

HDF5.h5s_create_simple

HDF5.h5s_get_simple_extent_dims

HDF5.h5s_get_simple_extent_ndims

HDF5.h5s_get_simple_extent_type

HDF5.h5s_is_simple

HDF5.h5s_select_hyperslab

HDF5.h5t_array_create

HDF5.h5t_close

HDF5.h5t_commit

HDF5.h5t_committed

HDF5.h5t_copy

HDF5.h5t_create

HDF5.h5t_equal

HDF5.h5t_get_array_dims

HDF5.h5t_get_array_ndims

HDF5.h5t_get_class

HDF5.h5t_get_cset

HDF5.h5t_get_member_class

HDF5.h5t_get_member_index

HDF5.h5t_get_member_name

HDF5.h5t_get_member_offset

HDF5.h5t_get_member_type

HDF5.h5t_get_native_type

HDF5.h5t_get_nmembers

HDF5.h5t_get_sign

HDF5.h5t_get_size

HDF5.h5t_get_strpad

HDF5.h5t_get_super

HDF5.h5t_get_tag

HDF5.h5t_insert

HDF5.h5t_is_variable_str

HDF5.h5t_open

HDF5.h5t_set_cset

HDF5.h5t_set_precision

HDF5.h5t_set_size

HDF5.h5t_vlen_create

HDF5.h5write

HDF5.h5writeattr

HDF5.has

HDF5.hash

HDF5.hdf5_obj_open

HDF5.hdf5_prop_get_set

HDF5.hdf5_to_julia

HDF5.hdf5_to_julia_eltype

HDF5.hdf5_type_id

HDF5.hdf5_type_map

HDF5.hdf5array

HDF5.heuristic_chunk

HDF5.hf5start_swmr_write

HDF5.hiding_errors

HDF5.hyperslab

HDF5.include

HDF5.info

HDF5.init_libhdf5

HDF5.iscontiguous

HDF5.ishdf5 — Function.

ishdf5(name::AbstractString)

Returns true if name is a path to a valid hdf5 file, false otherwise.

HDF5.ismmappable

HDF5.isnull

HDF5.libhdf5

HDF5.libhdf5_hl

HDF5.libhdf5handle

HDF5.libversion

HDF5.mpihandles

HDF5.name

HDF5.o_copy

HDF5.o_delete

HDF5.o_open

HDF5.objinfo

HDF5.p2a

HDF5.p_create

HDF5.read_const

HDF5.read_row

HDF5.readarray

HDF5.readmmap

HDF5.refresh

HDF5.register_blosc

HDF5.rehash!

HDF5.root

HDF5.set_chunk

HDF5.set_dims! — Function.

set_dims!(dset::HDF5Dataset, new_dims::Dims)

Change the current dimensions of a dataset to new_dims, limited by max_dims = get_dims(dset)[2]. Reduction is possible and leads to loss of truncated data.

HDF5.split1

HDF5.start_swmr_write — Function.

start_swmr_write(h5::HDF5File)

Start Single Reader Multiple Writer (SWMR) writing mode. See SWMR documentation.

HDF5.stringtype

HDF5.t2p

HDF5.t_commit

HDF5.t_create

HDF5.t_open

HDF5.unpad

HDF5.vlen_get_buf_size

HDF5.vlenpack

HDF5.writearray