Docstrings

ExcelReaders.ExcelErrorCell — Type.

ExcelErrorCell

An Excel cell that has an Excel error.

You cannot create $ExcelErrorCell$ objects, they are returned if a cell in an Excel file has an Excel error.

ExcelReaders.ExcelFile — Type.

ExcelFile

A handle to an open Excel file.

You can create an instance of an $ExcelFile$ by calling $openxl$.

ExcelReaders.ExcelReaders — Module.

ExcelReaders

Synopsis

ExcelReaders allows you to read data from Excel files.

Exported functions

$openxl$: Open an Excel file for repeated reads.

$readxl$: Read data from a specified range in an Excel file.

$readxlsheet$: Read data from a specified sheet in an Excel file.

Exported types

$ExcelFile$: A handle to an Excel file that was opened with $openxl$.

$ExcelErrorCell$: Value returned for cells with an Excel error code.

Further information

The $ExcelReaders.tutorial$ help topic provides a tutorial to the package.

The package homepage is https://github.com/davidanthoff/ExcelReaders.jl. Please report any issues with the package there.

ExcelReaders.__init__

ExcelReaders.colnum

ExcelReaders.convert_args_to_row_col

ExcelReaders.convert_ref_to_sheet_row_col

ExcelReaders.eval

ExcelReaders.get_cell_value

ExcelReaders.include

ExcelReaders.openxl — Function.

openxl(filename)

Open the Excel file $filename$ and return an $ExcelFile$ handle.

The returned $ExcelFile$ handle can later be passed as the first argument to $readxl$ or $readxslsheet$ to read from that file. If you will call either of those functions more than once, performance will be better if you open the file only once with $openxl$.

Example

f = openxl("filename.xlsx")
data = readxl(f, "Sheet1!A1:C4")

ExcelReaders.readxl

ExcelReaders.readxl_internal

ExcelReaders.readxlnames

ExcelReaders.readxlrange

ExcelReaders.readxlsheet

ExcelReaders.tutorial — Constant.

ExcelReaders.tutorial

Basic usage

The most basic usage is this:

using ExcelReaders
data = readxl("Filename.xlsx", "Sheet1!A1:C4")

This will return an array with all the data in the cell range A1 to C4 on Sheet1 in the Excel file Filename.xlsx.

If you expect to read multiple ranges from the same Excel file you can get much better performance by opening the Excel file only once:

using ExcelReaders
f = openxl("Filename.xlsx")
data1 = readxl(f, "Sheet1!A1:C4")
data2 = readxl(f, "Sheet2!B4:F10")

Reading a whole sheet

The $readxlsheet$ function reads complete Excel sheets, without a need to specify precise range information. The most basic usage is

using ExcelReaders
data = readxlsheet("Filename.xlsx", "Sheet1")

This will read all content on Sheet1 in the file Filename.xlsx. Eventual blank rows and columns at the top and left are skipped. $readxlsheet$ takes a number of optional keyword arguments:

$skipstartrows$ accepts either $:blanks$ (default) or a positive integer.

With $:blank$ any empty initial rows are skipped. An integer skips as many rows as specified.

$skipstartcols$ accepts either $:blanks$ (default) or a positive integer.

With $:blank$ any empty initial columns are skipped. An integer skips as many columns as specified.

$nrows$ accepts either $:all$ (default) or a positive integer. With

$:all$, all rows (except skipped ones) are read. An integer specifies the exact number of rows to be read.

$ncols$ accepts either $:all$ (default) or a postiive integer. With $:all$,

all columns (except skipped ones) are read. An integer specifies the exact number of columns to be read.

$readxlsheet$ also accepts an $ExcelFile$ (as obtained from $openxl$) as its first argument.

ExcelReaders.xlrd