Docstrings

StatsPlots.@animate

StatsPlots.@colorant_str

StatsPlots.@df — Macro.

`@df d x`

Convert every symbol in the expression x with the respective column in d if it exists.

If you want to avoid replacing the symbol, escape it with ^.

NA values are replaced with NaN for columns of Float64 and "" or Symbol() for strings and symbols respectively.

x can be either a plot command or a block of plot commands.

`@df x`

Curried version of @df d x. Outputs an anonymous function d -> @df d x.

StatsPlots.@gif

StatsPlots.@layout

StatsPlots.@recipe

StatsPlots.@series

StatsPlots.@shorthands

StatsPlots.@userplot

StatsPlots.ABGR

StatsPlots.ADIN99

StatsPlots.ADIN99d

StatsPlots.ADIN99o

StatsPlots.AGray

StatsPlots.AGray32

StatsPlots.AHSI

StatsPlots.AHSL

StatsPlots.AHSV

StatsPlots.ALCHab

StatsPlots.ALCHuv

StatsPlots.ALMS

StatsPlots.ALab

StatsPlots.ALuv

StatsPlots.ARGB

StatsPlots.ARGB32

StatsPlots.AXYZ

StatsPlots.AYCbCr

StatsPlots.AYIQ

StatsPlots.AbstractBackend

StatsPlots.AbstractGray

StatsPlots.AbstractLayout

StatsPlots.AbstractPlot

StatsPlots.AbstractRGB

StatsPlots.AlphaColor

StatsPlots.AndrewsPlot

StatsPlots.Animation

StatsPlots.AxyY

StatsPlots.BGR

StatsPlots.BGRA

StatsPlots.BezierCurve

StatsPlots.CIE1931JV_CMF

StatsPlots.CIE1931J_CMF

StatsPlots.CIE1931_CMF

StatsPlots.CIE1964_CMF

StatsPlots.Color

StatsPlots.Color3

StatsPlots.ColorAlpha

StatsPlots.ColorGradient

StatsPlots.ColorTypes

StatsPlots.Colorant

StatsPlots.ColorantNormed

StatsPlots.Colors

StatsPlots.CornerPlot

StatsPlots.CorrPlot

StatsPlots.DE_2000

StatsPlots.DE_94

StatsPlots.DE_AB

StatsPlots.DE_BFD

StatsPlots.DE_CMC

StatsPlots.DE_DIN99

StatsPlots.DE_DIN99d

StatsPlots.DE_DIN99o

StatsPlots.DE_JPC79

StatsPlots.DIN99

StatsPlots.DIN99A

StatsPlots.DIN99d

StatsPlots.DIN99dA

StatsPlots.DIN99o

StatsPlots.DIN99oA

StatsPlots.Formatted

StatsPlots.Fractional

StatsPlots.GR

StatsPlots.Gray

StatsPlots.Gray24

StatsPlots.GrayA

StatsPlots.GroupedBar

StatsPlots.HSB

StatsPlots.HSI

StatsPlots.HSIA

StatsPlots.HSL

StatsPlots.HSLA

StatsPlots.HSV

StatsPlots.HSVA

StatsPlots.KW

StatsPlots.LCHab

StatsPlots.LCHabA

StatsPlots.LCHuv

StatsPlots.LCHuvA

StatsPlots.LMS

StatsPlots.LMSA

StatsPlots.Lab

StatsPlots.LabA

StatsPlots.Luv

StatsPlots.LuvA

StatsPlots.MSC

StatsPlots.OHLC

StatsPlots.PlotThemes

StatsPlots.PlotUtils

StatsPlots.Plots

StatsPlots.QQNorm

StatsPlots.QQPlot

StatsPlots.RGB

StatsPlots.RGB1

StatsPlots.RGB24

StatsPlots.RGB4

StatsPlots.RGBA

StatsPlots.RecipeData

StatsPlots.RecipesBase

StatsPlots.Segments

StatsPlots.Shape

StatsPlots.StatsPlots

StatsPlots.Surface

StatsPlots.Transparent3

StatsPlots.TransparentColor

StatsPlots.TransparentGray

StatsPlots.TransparentRGB

StatsPlots.XYZ

StatsPlots.XYZA

StatsPlots.YCbCr

StatsPlots.YCbCrA

StatsPlots.YIQ

StatsPlots.YIQA

StatsPlots._arg2string

StatsPlots._argnames

StatsPlots._violin_warned

StatsPlots.adapted_grid

StatsPlots.add_backend

StatsPlots.add_label

StatsPlots.add_theme

StatsPlots.aliases

StatsPlots.alpha

StatsPlots.alphacolor

StatsPlots.andrewsplot — Function.

andrewsplot(args...; kw...)

Shows each row of an array (or table) as a line. The x argument specifies a grouping variable. This is a way to visualize structure in high-dimensional data. https://en.wikipedia.org/wiki/Andrews_plot #Examples

using RDatasets, StatsPlots
iris = dataset("datasets", "iris")
@df iris andrewsplot(:Species, cols(1:4))

StatsPlots.andrewsplot!

StatsPlots.animate

StatsPlots.annotate!

StatsPlots.arrow

StatsPlots.attr!

StatsPlots.backend

StatsPlots.backend_name

StatsPlots.backend_object

StatsPlots.backends

StatsPlots.bar

StatsPlots.bar!

StatsPlots.barh

StatsPlots.barh!

StatsPlots.barhist

StatsPlots.barhist!

StatsPlots.base_color_type

StatsPlots.base_colorant_type

StatsPlots.bbox

StatsPlots.bkfe — Function.

binned kernel function estimator

StatsPlots.blue

StatsPlots.boxplot

StatsPlots.boxplot!

StatsPlots.brush

StatsPlots.ccolor

StatsPlots.cdensity

StatsPlots.cdensity!

StatsPlots.center

StatsPlots.cgrad

StatsPlots.cgradients

StatsPlots.cie_color_match

StatsPlots.clibraries

StatsPlots.clibrary

StatsPlots.closeall

StatsPlots.color

StatsPlots.color_type

StatsPlots.coloralpha

StatsPlots.colordiff

StatsPlots.colormap

StatsPlots.colormatch

StatsPlots.column_names

StatsPlots.column_types

StatsPlots.comp1

StatsPlots.comp2

StatsPlots.comp3

StatsPlots.compute_name

StatsPlots.contour

StatsPlots.contour!

StatsPlots.contour3d

StatsPlots.contour3d!

StatsPlots.contourf

StatsPlots.contourf!

StatsPlots.convert_missing

StatsPlots.coords

StatsPlots.cornerplot

StatsPlots.cornerplot!

StatsPlots.corrplot

StatsPlots.corrplot!

StatsPlots.current

StatsPlots.curve_points

StatsPlots.curves

StatsPlots.curves!

StatsPlots.cvec

StatsPlots.dataviewer

StatsPlots.default

StatsPlots.default_cgrad

StatsPlots.default_range

StatsPlots.density

StatsPlots.density!

StatsPlots.deuteranopic

StatsPlots.df_helper

StatsPlots.distinguishable_colors

StatsPlots.diverging_palette

StatsPlots.ea_binnumber

StatsPlots.ea_histogram

StatsPlots.ea_histogram!

StatsPlots.eval

StatsPlots.extract_columns_from_iterabletable

StatsPlots.font

StatsPlots.frame

StatsPlots.gamutmax

StatsPlots.gamutmin

StatsPlots.get_col_from_dict

StatsPlots.get_color_palette

StatsPlots.gif

StatsPlots.gr

StatsPlots.gray

StatsPlots.green

StatsPlots.grid

StatsPlots.grouped_xy

StatsPlots.groupedbar

StatsPlots.groupedbar!

StatsPlots.gui

StatsPlots.hdf5

StatsPlots.heatmap

StatsPlots.heatmap!

StatsPlots.hex

StatsPlots.hexbin

StatsPlots.hexbin!

StatsPlots.histogram

StatsPlots.histogram!

StatsPlots.histogram2d

StatsPlots.histogram2d!

StatsPlots.hline

StatsPlots.hline!

StatsPlots.hspan

StatsPlots.hspan!

StatsPlots.include

StatsPlots.inline

StatsPlots.insert_kw!

StatsPlots.inspectdr

StatsPlots.invisible

StatsPlots.isdark

StatsPlots.iter_segments

StatsPlots.linbin

StatsPlots.loc

StatsPlots.mapc

StatsPlots.mapreducec

StatsPlots.marginalhist

StatsPlots.marginalhist!

StatsPlots.mov

StatsPlots.mp4

StatsPlots.not_kw

StatsPlots.notch_width

StatsPlots.ohlc

StatsPlots.ohlc!

StatsPlots.optimize_datetime_ticks

StatsPlots.optimize_ticks

StatsPlots.palette

StatsPlots.parse_iterabletable_call!

StatsPlots.path3d

StatsPlots.path3d!

StatsPlots.pgfplots

StatsPlots.pie

StatsPlots.pie!

StatsPlots.plot

StatsPlots.plot!

StatsPlots.plot3d

StatsPlots.plot3d!

StatsPlots.plot_color

StatsPlots.plotarea

StatsPlots.plotattr

StatsPlots.plotly

StatsPlots.plotlyjs

StatsPlots.plots_heatmap

StatsPlots.plots_heatmap!

StatsPlots.png

StatsPlots.portfoliocomposition

StatsPlots.portfoliocomposition!

StatsPlots.protanopic

StatsPlots.pyplot

StatsPlots.qqnorm

StatsPlots.qqnorm!

StatsPlots.qqplot

StatsPlots.qqplot!

StatsPlots.quiver

StatsPlots.quiver!

StatsPlots.red

StatsPlots.reducec

StatsPlots.rgb_string

StatsPlots.rgba_string

StatsPlots.rotate

StatsPlots.rotate!

StatsPlots.savefig

StatsPlots.scatter

StatsPlots.scatter!

StatsPlots.scatter3d

StatsPlots.scatter3d!

StatsPlots.scatterhist

StatsPlots.scatterhist!

StatsPlots.sequential_palette

StatsPlots.set_theme

StatsPlots.shape_coords

StatsPlots.showgradient

StatsPlots.showgradient!

StatsPlots.showlibrary

StatsPlots.showlibrary!

StatsPlots.showtheme

StatsPlots.showtheme!

StatsPlots.spy

StatsPlots.spy!

StatsPlots.stephist

StatsPlots.stephist!

StatsPlots.sticks

StatsPlots.sticks!

StatsPlots.stringify

StatsPlots.stroke

StatsPlots.surface

StatsPlots.surface!

StatsPlots.test_examples

StatsPlots.testhist

StatsPlots.testhist!

StatsPlots.text

StatsPlots.theme

StatsPlots.title!

StatsPlots.translate

StatsPlots.translate!

StatsPlots.treepositions

StatsPlots.tritanopic

StatsPlots.twinx

StatsPlots.unicodeplots

StatsPlots.update_ticks_guides

StatsPlots.violin

StatsPlots.violin!

StatsPlots.violin_coords

StatsPlots.vline

StatsPlots.vline!

StatsPlots.vspan

StatsPlots.vspan!

StatsPlots.wand_bins — Function.

Returns optimal histogram bin widths in accordance to Wand (1995)'s criterion'

StatsPlots.weighted_color_mean

StatsPlots.whitebalance

StatsPlots.wireframe

StatsPlots.wireframe!

StatsPlots.with

StatsPlots.wrap

StatsPlots.xaxis!

StatsPlots.xflip!

StatsPlots.xgrid!

StatsPlots.xlabel!

StatsPlots.xlims

StatsPlots.xlims!

StatsPlots.xticks!

StatsPlots.xyY

StatsPlots.xyYA

StatsPlots.yaxis!

StatsPlots.yflip!

StatsPlots.ygrid!

StatsPlots.ylabel!

StatsPlots.ylims

StatsPlots.ylims!

StatsPlots.yticks!

StatsPlots.yz_args

StatsPlots.zlims

StatsPlots.zlims!