Docstrings

OffsetArrays.@unsafe

OffsetArrays.ArrayInitializer

OffsetArrays.OffsetArray

OffsetArrays.OffsetArrays

OffsetArrays.OffsetAxis

OffsetArrays.OffsetRange

OffsetArrays.OffsetVector

OffsetArrays._axes

OffsetArrays._slice

OffsetArrays._unslice

OffsetArrays.eval

OffsetArrays.include

OffsetArrays.indexlength

OffsetArrays.indexoffset

OffsetArrays.no_offset_view — Function.

no_offset_view(A)

Return an AbstractArray that shares structure and has the same type and size as the argument, but has 1-based indexing. May just return the argument when applicable. Not exported.

The default implementation uses OffsetArrays, but other types should use something more specific to remove a level of indirection when applicable.

julia> O = OffsetArray(A, 0:1, -1:1)
OffsetArray(::Array{Int64,2}, 0:1, -1:1) with eltype Int64 with indices 0:1×-1:1:
 1  3  5
 2  4  6

julia> OffsetArrays.no_offset_view(O)[1,1] = -9
-9

julia> A
2×3 Array{Int64,2}:
 -9  3  5
  2  4  6

OffsetArrays.offset

OffsetArrays.parenttype

OffsetArrays.printindices