Docstrings

IndexedTables.@cols

IndexedTables.AbstractIndexedTable

IndexedTables.AbstractNDSparse

IndexedTables.All — Type.

All(cols::Union{Symbol, Int}...)

Select the union of the selections in cols. If cols == (), select all columns.

Examples

t = table([1,1,2,2], [1,2,1,2], [1,2,3,4], [0, 0, 0, 0], names=[:a,:b,:c,:d])
select(t, All(:a, (:b, :c)))
select(t, All())

IndexedTables.ApplyColwise

IndexedTables.Between — Type.

Between(first, last)

Select the columns between first and last.

Examples

t = table([1,1,2,2], [1,2,1,2], 1:4, 'a':'d', names=[:a,:b,:c,:d])
select(t, Between(:b, :d))

IndexedTables.ColDict — Type.

d = ColDict(t)

Create a mutable dictionary of columns in t.

To get the immutable iterator of the same type as t call d[]

IndexedTables.Columns

IndexedTables.Dataset

IndexedTables.DimName

IndexedTables.GroupBy

IndexedTables.GroupReduce

IndexedTables.IndexedTable — Type.

A tabular data structure that extends Columns. Create an IndexedTable with the table function.

IndexedTables.IndexedTables

IndexedTables.Keys — Type.

Keys()

Select the primary keys.

Examples

t = table([1,1,2,2], [1,2,1,2], [1,2,3,4], names=[:a,:b,:c], pkey = (:a, :b))
select(t, Keys())

IndexedTables.NDSparse — Type.

NDSparse(columns...; names=Symbol[...], kwargs...)

Construct an NDSparse array from columns. The last argument is the data column, and the rest are index columns. The names keyword argument optionally specifies names for the index columns (dimensions).

IndexedTables.NextTable

IndexedTables.Not — Type.

Not(cols::Union{Symbol, Int}...)

Select the complementary of the selection in cols. Not can accept several arguments, in which case it returns the complementary of the union of the selections.

Examples

t = table([1,1,2,2], [1,2,1,2], [1,2,3,4], names=[:a,:b,:c], pkey = (:a, :b))
select(t, Not(:a))
select(t, Not(:a, (:a, :b)))

IndexedTables.Perm — Type.

A permutation

Fields:

columns: The columns being indexed as a vector of integers (column numbers)

perm: the permutation - an array or iterator which has the sorted permutation

IndexedTables.SpecialSelector

IndexedTables.Tup

IndexedTables._append!

IndexedTables._apply

IndexedTables._apply_with_key

IndexedTables._bcast_loop!

IndexedTables._broadcast!

IndexedTables._broadcast_trailing!

IndexedTables._colindex

IndexedTables._cols

IndexedTables._cols_tuple

IndexedTables._convert

IndexedTables._dep_message_NextTable

IndexedTables._dep_message_dropna

IndexedTables._flatten!

IndexedTables._fwd

IndexedTables._getindex

IndexedTables._getindex_scalar

IndexedTables._impl

IndexedTables._in

IndexedTables._ismissing

IndexedTables._join!

IndexedTables._map

IndexedTables._map_params

IndexedTables._mapslices_itable!

IndexedTables._mapslices_itable_singleton!

IndexedTables._mapslices_scalar!

IndexedTables._merge

IndexedTables._merge!

IndexedTables._promote_op

IndexedTables._push!

IndexedTables._reduce

IndexedTables._reduce_select

IndexedTables._reduce_select_init

IndexedTables._row_in

IndexedTables._setindex!

IndexedTables._setindex_scalar!

IndexedTables._sizehint!

IndexedTables._strip_pair

IndexedTables._tuple_type_head

IndexedTables._tuple_type_tail

IndexedTables.addname

IndexedTables.aggregate! — Function.

aggregate!(f::Function, arr::NDSparse)

Combine adjacent rows with equal indices using the given 2-argument reduction function, in place.

IndexedTables.antijoin

IndexedTables.append_n!

IndexedTables.arrayof — Function.

arrayof(T)

Returns the type of Columns or Vector suitable to store values of type T. Nested tuples beget nested Columns.

IndexedTables.asofjoin — Function.

asofjoin(left::NDSparse, right::NDSparse)

Join rows from left with the "most recent" value from right.

Example

using Dates
akey1 = ["A", "A", "B", "B"]
akey2 = [Date(2017,11,11), Date(2017,11,12), Date(2017,11,11), Date(2017,11,12)]
avals = collect(1:4)

bkey1 = ["A", "A", "B", "B"]
bkey2 = [Date(2017,11,12), Date(2017,11,13), Date(2017,11,10), Date(2017,11,13)]
bvals = collect(5:8)

a = ndsparse((akey1, akey2), avals)
b = ndsparse((bkey1, bkey2), bvals)

asofjoin(a, b)

IndexedTables.astuple

IndexedTables.best_perm_estimate — Function.

Returns: (n, perm) where n is the number of columns in the beginning of cols, perm is one possible permutation of those first n columns.

IndexedTables.cacheperm!

IndexedTables.canonname

IndexedTables.cmpelts

IndexedTables.colindex

IndexedTables.collect_columns — Function.

collect_columns(itr)

Collect an iterable as a Columns object if it iterates Tuples or NamedTuples, as a normal Array otherwise.

Examples

s = [(1,2), (3,4)]
collect_columns(s)

s2 = Iterators.filter(isodd, 1:8)
collect_columns(s2)

IndexedTables.collect_columns_flattened

IndexedTables.collect_columns_flattened!

IndexedTables.collect_empty_columns

IndexedTables.collect_to_columns!

IndexedTables.collectiontype

IndexedTables.colname

IndexedTables.colnames — Function.

colnames(itr)

Returns the names of the "columns" in itr.

Examples:

colnames(1:3)
colnames(Columns([1,2,3], [3,4,5]))
colnames(table([1,2,3], [3,4,5]))
colnames(Columns(x=[1,2,3], y=[3,4,5]))
colnames(table([1,2,3], [3,4,5], names=[:x,:y]))
colnames(ndsparse(Columns(x=[1,2,3]), Columns(y=[3,4,5])))
colnames(ndsparse(Columns(x=[1,2,3]), [3,4,5]))
colnames(ndsparse(Columns(x=[1,2,3]), [3,4,5]))
colnames(ndsparse(Columns([1,2,3], [4,5,6]), Columns(x=[6,7,8])))
colnames(ndsparse(Columns(x=[1,2,3]), Columns([3,4,5],[6,7,8])))

IndexedTables.column

IndexedTables.columns — Function.

columns(itr, select::Selection = All())

Select one or more columns from an iterable of rows as a tuple of vectors.

select specifies which columns to select. Refer to the select function for the available selection options and syntax.

itr can be NDSparse, Columns, AbstractVector, or their distributed counterparts.

Examples

t = table(1:2, 3:4; names = [:x, :y])

columns(t)
columns(t, :x)
columns(t, (:x,))
columns(t, (:y, :x => -))

IndexedTables.colwise_group_fast

IndexedTables.compact_mem

IndexedTables.concat_cols

IndexedTables.concat_tup

IndexedTables.concat_tup_type

IndexedTables.convertdim — Function.

convertdim(x::NDSparse, d::DimName, xlate; agg::Function, vecagg::Function, name)

Apply function or dictionary xlate to each index in the specified dimension. If the mapping is many-to-one, agg or vecagg is used to aggregate the results. If agg is passed, it is used as a 2-argument reduction function over the data. If vecagg is passed, it is used as a vector-to-scalar function to aggregate the data. name optionally specifies a new name for the translated dimension.

IndexedTables.convertkey

IndexedTables.convertmissing — Function.

convertmissing(tbl, missingtype)

Convert the missing value representation in tbl to be of type missingtype.

Example

using IndexedTables, DataValues
t = table([1,2,missing], [1,missing,3])
IndexedTables.convertmissing(t, DataValue)

IndexedTables.copyelt!

IndexedTables.copyrow!

IndexedTables.count_overlap

IndexedTables.dedup_names

IndexedTables.default_initializer

IndexedTables.dimlabels — Function.

dimlabels(t::NDSparse)

Returns an array of integers or symbols giving the labels for the dimensions of t. ndims(t) == length(dimlabels(t)).

IndexedTables.dropmissing — Function.

dropmissing(t        )
dropmissing(t, select)

Drop rows of table t which contain missing values (either Missing or DataValue), optionally only using the columns in select. Column types will be converted to non-missing types. For example:

Vector{Union{Int, Missing}} -> Vector{Int}

DataValueArray{Int} -> Vector{Int}

Example

t = table([0.1,0.5,missing,0.7], [2,missing,4,5], [missing,6,missing,7], names=[:t,:x,:y])
dropmissing(t)
dropmissing(t, (:t, :x))

IndexedTables.dropna — Function.

dropmissing(t        )
dropmissing(t, select)

Drop rows of table t which contain missing values (either Missing or DataValue), optionally only using the columns in select. Column types will be converted to non-missing types. For example:

Vector{Union{Int, Missing}} -> Vector{Int}

DataValueArray{Int} -> Vector{Int}

Example

t = table([0.1,0.5,missing,0.7], [2,missing,4,5], [missing,6,missing,7], names=[:t,:x,:y])
dropmissing(t)
dropmissing(t, (:t, :x))

IndexedTables.eltypes

IndexedTables.eval

IndexedTables.excludecols — Function.

excludecols(itr, cols) -> Tuple of Int

Names of all columns in itr except cols. itr can be any of Table, NDSparse, Columns, or AbstractVector

Examples

using IndexedTables: excludecols

t = table([2,1],[1,3],[4,5], names=[:x,:y,:z], pkey=(1,2))

excludecols(t, (:x,))
excludecols(t, (2,))
excludecols(t, pkeynames(t))
excludecols([1,2,3], (1,))

IndexedTables.fastmap

IndexedTables.fieldindex

IndexedTables.fieldstupletype

IndexedTables.fieldtypes

IndexedTables.filt_by_col!

IndexedTables.find_corresponding

IndexedTables.flatten — Function.

flatten(t::Table, col=length(columns(t)))

Flatten col column which may contain a vector of vectors while repeating the other fields. If column argument is not provided, default to last column.

Examples:

x = table([1,2], [[3,4], [5,6]], names=[:x, :y])
flatten(x, 2)

t1 = table([3,4],[5,6], names=[:a,:b])
t2 = table([7,8], [9,10], names=[:a,:b])
x = table([1,2], [t1, t2], names=[:x, :y]);
flatten(x, :y)

IndexedTables.flush! — Function.

flush!(arr::NDSparse)

Commit queued assignment operations, by sorting and merging the internal temporary buffer.

IndexedTables.foreach

IndexedTables.getsubfields

IndexedTables.groupby — Function.

groupby(f, t, by = pkeynames(t); select, flatten=false)

Apply f to the select-ed columns (see select) in groups defined by the unique values of by.

If f returns a vector, split it into multiple columns with flatten = true.

Examples

using Statistics

t=table([1,1,1,2,2,2], [1,1,2,2,1,1], [1,2,3,4,5,6], names=[:x,:y,:z])

groupby(mean, t, :x, select=:z)
groupby(identity, t, (:x, :y), select=:z)
groupby(mean, t, (:x, :y), select=:z)

groupby((mean, std, var), t, :y, select=:z)
groupby((q25=z->quantile(z, 0.25), q50=median, q75=z->quantile(z, 0.75)), t, :y, select=:z)

# apply different aggregation functions to different columns
groupby((ymean = :y => mean, zmean = :z => mean), t, :x)

IndexedTables.groupjoin — Function.

groupjoin(left, right; kw...)
groupjoin(f, left, right; kw...)

Join left and right creating groups of values with matching keys.

For keyword argument options, see join.

Examples

l = table([1,1,1,2], [1,2,2,1], [1,2,3,4], names=[:a,:b,:c], pkey=(:a, :b))
r = table([0,1,1,2], [1,2,2,1], [1,2,3,4], names=[:a,:b,:d], pkey=(:a, :b))

groupjoin(l, r)
groupjoin(l, r; how = :left)
groupjoin(l, r; how = :outer)
groupjoin(l, r; how = :anti)

IndexedTables.groupreduce — Function.

groupreduce(f, t, by = pkeynames(t); select)

Calculate a reduce operation f over table t on groups defined by the values in selection by. The result is put in a table keyed by the unique by values.

Examples

t = table([1,1,1,2,2,2], 1:6, names=[:x, :y])
groupreduce(+,        t, :x; select = :y)
groupreduce((sum=+,), t, :x; select = :y)  # change output column name to :sum

t2 = table([1,1,1,2,2,2], [1,1,2,2,3,3], 1:6, names = [:x, :y, :z])
groupreduce(+, t2, (:x, :y), select = :z)

# different reducers for different columns
groupreduce((sumy = :y => +, sumz = :z => +), t2, :x)

IndexedTables.grow_to_columns!

IndexedTables.hascolumns

IndexedTables.include

IndexedTables.init_first

IndexedTables.init_func

IndexedTables.init_funcs

IndexedTables.init_inputs

IndexedTables.init_join_output

IndexedTables.innerjoin

IndexedTables.insertafter!

IndexedTables.insertbefore!

IndexedTables.insertcol — Function.

insertcol(t, position::Integer, name, x)

Insert a column x named name at position. Returns a new table.

Example

t = table([0.01, 0.05], [2,1], [3,4], names=[:t, :x, :y], pkey=:t)
insertcol(t, 2, :w, [0,1])

IndexedTables.insertcolafter — Function.

insertcolafter(t, after, name, col)

Insert a column col named name after after. Returns a new table.

Example

t = table([0.01, 0.05], [2,1], [3,4], names=[:t, :x, :y], pkey=:t)
insertcolafter(t, :t, :w, [0,1])

IndexedTables.insertcolbefore — Function.

insertcolbefore(t, before, name, col)

Insert a column col named name before before. Returns a new table.

Example

t = table([0.01, 0.05], [2,1], [3,4], names=[:t, :x, :y], pkey=:t)
insertcolbefore(t, :x, :w, [0,1])

IndexedTables.is_approx_uniqs_less_than

IndexedTables.isconstrange

IndexedTables.ismissingtype

IndexedTables.isshared

IndexedTables.ith_all

IndexedTables.joinequalblock

IndexedTables.left

IndexedTables.leftgroupjoin

IndexedTables.leftjoin

IndexedTables.lowerselection

IndexedTables.map_params

IndexedTables.map_rows — Function.

map_rows(f, c...)

Transform collection c by applying f to each element. For multiple collection arguments, apply f elementwise. Collect output as Columns if f returns Tuples or NamedTuples with constant fields, as Array otherwise.

Examples

map_rows(i -> (exp = exp(i), log = log(i)), 1:5)

IndexedTables.mapped_type

IndexedTables.match_indices

IndexedTables.missing_indxs

IndexedTables.missing_instance

IndexedTables.missingtype2type

IndexedTables.namedtuple

IndexedTables.naturalgroupjoin

IndexedTables.naturaljoin

IndexedTables.ncols — Function.

ncols(itr)

Returns the number of columns in itr.

Examples

ncols([1,2,3]) == 1
ncols(rows(([1,2,3],[4,5,6]))) == 2

IndexedTables.ndsparse — Function.

ndsparse(keys, values; kw...)

Construct an NDSparse array with the given keys and values columns. On construction, the keys and data are sorted in lexicographic order of the keys.

Keyword Argument Options:

agg = nothing – Function to aggregate values with duplicate keys.

presorted = false – Are the key columns already sorted?

copy = true – Should the columns in keys and values be copied?

chunks = nothing – Provide an integer to distribute data into chunks chunks.

A good choice is nworkers() (after using Distributed)

See also: distribute

Examples:

x = ndsparse(["a","b"], [3,4])
keys(x)
values(x)
x["a"]

# Dimensions are named if constructed with a named tuple of columns 
x = ndsparse((index = 1:10,), rand(10))
x[1]

# Multiple dimensions by passing a (named) tuple of columns
x = ndsparse((x = 1:10, y = 1:2:20), rand(10))
x[1, 1]

# Value columns can also have names via named tuples
x = ndsparse(1:10, (x=rand(10), y=rand(10)))

IndexedTables.nicename

IndexedTables.nullrow

IndexedTables.opt_vcat

IndexedTables.outergroupjoin

IndexedTables.outerjoin

IndexedTables.outvec

IndexedTables.permcache

IndexedTables.pkeynames — Function.

pkeynames(t::Table)

Names of the primary key columns in t.

Examples

t = table([1,2], [3,4]);
pkeynames(t)

t = table([1,2], [3,4], pkey=1);
pkeynames(t)

t = table([2,1],[1,3],[4,5], names=[:x,:y,:z], pkey=(1,2));
pkeynames(t)

pkeynames(t::NDSparse)

Names of the primary key columns in t.

Example

x = ndsparse([1,2],[3,4])
pkeynames(x)

x = ndsparse((x=1:10, y=1:2:20), rand(10))
pkeynames(x)

IndexedTables.pkeys — Function.

pkeys(itr::IndexedTable)

Primary keys of the table. If Table doesn't have any designated primary key columns (constructed without pkey argument) then a default key of tuples (1,):(n,) is generated.

Example

a = table(["a","b"], [3,4]) # no pkey
pkeys(a)

a = table(["a","b"], [3,4], pkey=1)
pkeys(a)

IndexedTables.popcol — Function.

popcol(t, cols...)

Remove the column(s) cols from the table. Returns a new table.

Example

t = table([0.01, 0.05], [2,1], [3,4], names=[:t, :x, :y], pkey=:t)
popcol(t, :x)

IndexedTables.primaryperm

IndexedTables.product

IndexedTables.promote_union

IndexedTables.promoted_similar

IndexedTables.pushcol — Function.

pushcol(t, name, x)

Push a column x to the end of the table. name is the name for the new column. Returns a new table.

pushcol(t, map::Pair...)

Push many columns at a time.

Example

t = table([0.01, 0.05], [2,1], [3,4], names=[:t, :x, :y], pkey=:t)
pushcol(t, :z, [1//2, 3//4])
pushcol(t, :z => [1//2, 3//4])

IndexedTables.pushrow!

IndexedTables.range_estimate

IndexedTables.reduced_type

IndexedTables.reducedim_vec — Function.

reducedim_vec(f::Function, arr::NDSparse, dims)

Like reduce, except uses a function mapping a vector of values to a scalar instead of a 2-argument scalar function.

IndexedTables.reindex — Function.

reindex(t::IndexedTable, by)
reindex(t::IndexedTable, by, select)

Reindex table t with new primary key by, optionally keeping a subset of columns via select. For NDSparse, use selectkeys.

Example

t = table([2,1],[1,3],[4,5], names=[:x,:y,:z], pkey=(1,2))

t2 = reindex(t, (:y, :z))

pkeynames(t2)

IndexedTables.rename!

IndexedTables.renamecol — Function.

renamecol(t, col, newname)

Set newname as the new name for column col in t. Returns a new table.

renamecol(t, map::Pair...)

Rename multiple columns at a time.

Example

t = table([0.01, 0.05], [2,1], names=[:t, :x])
renamecol(t, :t, :time)

IndexedTables.right

IndexedTables.row_asof

IndexedTables.row_in

IndexedTables.rowcmp

IndexedTables.roweq

IndexedTables.rowless

IndexedTables.rows — Function.

rows(itr, select = All())

Select one or more fields from an iterable of rows as a vector of their values. Refer to the select function for selection options and syntax.

itr can be NDSparse, Columns, AbstractVector, or their distributed counterparts.

Examples

t = table([1,2],[3,4], names=[:x,:y])
rows(t)
rows(t, :x)
rows(t, (:x,))
rows(t, (:y, :x => -))

IndexedTables.select — Function.

select(t::Table, which::Selection)

Select all or a subset of columns, or a single column from the table.

Selection is a type union of many types that can select from a table. It can be:

Integer – returns the column at this position.

Symbol – returns the column with this name.

Pair{Selection => Function} – selects and maps a function over the selection, returns the result.

AbstractArray – returns the array itself. This must be the same length as the table.

Tuple of Selection – returns a table containing a column for every selector in the tuple. The tuple may also contain the type Pair{Symbol, Selection}, which the selection a name. The most useful form of this when introducing a new column.

Regex – returns the columns with names that match the regular expression.

Type – returns columns with elements of the given type.

Examples:

t = table(1:10, randn(10), rand(Bool, 10); names = [:x, :y, :z])

# select the :x vector
select(t, 1)
select(t, :x)

# map a function to the :y vector
select(t, 2 => abs)
select(t, :y => x -> x > 0 ? x : -x)

# select the table of :x and :z
select(t, (:x, :z))
select(t, r"(x|z)")

# map a function to the table of :x and :y
select(t, (:x, :y) => row -> row[1] + row[2])
select(t, (1, :y) => row -> row.x + row.y)

IndexedTables.selectkeys — Function.

selectkeys(x::NDSparse, sel)

Return an NDSparse with a subset of keys.

IndexedTables.selectvalues — Function.

selectvalues(x::NDSparse, sel)

Return an NDSparse with a subset of values

IndexedTables.set!

IndexedTables.set_show_compact!

IndexedTables.setcol — Function.

setcol(t::Table, col::Union{Symbol, Int}, x::Selection)

Sets a x as the column identified by col. Returns a new table.

setcol(t::Table, map::Pair{}...)

Set many columns at a time.

Examples:

t = table([1,2], [3,4], names=[:x, :y])

# change second column to [5,6]
setcol(t, 2 => [5,6])
setcol(t, :y , :y => x -> x + 2)

# add [5,6] as column :z 
setcol(t, :z => 5:6)
setcol(t, :z, :y => x -> x + 2)

# replacing the primary key results in a re-sorted copy
t = table([0.01, 0.05], [1,2], [3,4], names=[:t, :x, :y], pkey=:t)
t2 = setcol(t, :t, [0.1,0.05])

IndexedTables.show_compact_when_wide

IndexedTables.showmeta

IndexedTables.showtable

IndexedTables.sort_int_range_sub_by!

IndexedTables.sort_sub_by!

IndexedTables.sortperm_fast

IndexedTables.sortperm_int_range

IndexedTables.sortpermby

IndexedTables.stack — Function.

stack(t, by = pkeynames(t); select = Not(by), variable = :variable, value = :value)`

Reshape a table from the wide to the long format. Columns in by are kept as indexing columns. Columns in select are stacked. In addition to the id columns, two additional columns labeled variable and value are added, containing the column identifier and the stacked columns. See also unstack.

Examples

t = table(1:4, names = [:x], pkey=:x)
t = pushcol(t, :xsquare, :x => x -> x^2)
t = pushcol(t, :xcube  , :x => x -> x^3)

stack(t)

IndexedTables.strip_unionall

IndexedTables.strip_unionall_params

IndexedTables.subscriptprint

IndexedTables.subtable

IndexedTables.summarize — Function.

summarize(f, t, by = pkeynames(t); select = Not(by), stack = false, variable = :variable)

Apply summary functions column-wise to a table. Return a NamedTuple in the non-grouped case and a table in the grouped case. Use stack=true to stack results of the same summary function for different columns.

Examples

using Statistics

t = table([1, 2, 3], [1, 1, 1], names = [:x, :y])

summarize((mean, std), t)
summarize((m = mean, s = std), t)
summarize(mean, t; stack=true)
summarize((mean, std), t; select = :y)

IndexedTables.table — Function.

table(cols; kw...)

Create a table from a (named) tuple of AbstractVectors.

table(cols::AbstractVector...; names::Vector{Symbol}, kw...)

Create a table from the provided cols, optionally with names.

table(cols::Columns; kw...)

Construct a table from a vector of tuples. See rows and Columns.

table(t::Union{IndexedTable, NDSparse}; kw...)

Copy a Table or NDSparse to create a new table. The same primary keys as the input are used.

table(x; kw...)

Create an IndexedTable from any object x that follows the Tables.jl interface.

Keyword Argument Options:

pkey: select columns to sort by and be the primary key.

presorted = false: is the data pre-sorted by primary key columns?

copy = true: creates a copy of the input vectors if true. Irrelevant if chunks is specified.

chunks::Integer: distribute the table. Options are:

Int – (number of chunks) a safe bet is nworkers() after using Distributed.

Vector{Int} – Number of elements in each of the length(chunks) chunks.

Examples:

table(rand(10), rand(10), names = [:x, :y], pkey = :x)

table(rand(Bool, 20), rand(20), rand(20), pkey = [1,2])

table((x = 1:10, y = randn(10)))

table([(1,2), (3,4)])

IndexedTables.type2missingtype

IndexedTables.unstack — Function.

unstack(t, by = pkeynames(t); variable = :variable, value = :value)

Reshape a table from the long to the wide format. Columns in by are kept as indexing columns. Keyword arguments variable and value denote which column contains the column identifier and which the corresponding values. See also stack.

Examples

t = table(1:4, [1, 4, 9, 16], [1, 8, 27, 64], names = [:x, :xsquare, :xcube], pkey = :x);

long = stack(t)

unstack(long)

IndexedTables.unwrap

IndexedTables.update! — Function.

update!(f::Function, arr::NDSparse, indices...)

Replace data values x with f(x) at each location that matches the given indices.

IndexedTables.valuenames

IndexedTables.where — Function.

where(arr::NDSparse, indices...)

Returns an iterator over data items where the given indices match. Accepts the same index arguments as getindex.