Docstrings

PDMats.@check_argdims

PDMats.AbstractPDMat — Type.

The base type for positive definite matrices.

PDMats.CholType

PDMats.CholTypeSparse

PDMats.HAVE_CHOLMOD

PDMats.PDMat — Type.

Full positive definite matrix together with a Cholesky factorization object.

PDMats.PDMats

PDMats.PDSparseMat — Type.

Sparse positive definite matrix together with a Cholesky factorization object.

PDMats.PDiagMat — Type.

Positive definite diagonal matrix.

PDMats.ScalMat — Type.

Scaling matrix.

PDMats.X_A_Xt

PDMats.X_invA_Xt

PDMats.Xt_A_X

PDMats.Xt_invA_X

PDMats._adddiag

PDMats._adddiag!

PDMats._addscal!

PDMats._pdt

PDMats._rcopy!

PDMats.add!

PDMats.add_scal

PDMats.add_scal!

PDMats.chol_lower

PDMats.colwise_dot!

PDMats.colwise_sumsq!

PDMats.dim

PDMats.eval

PDMats.full

PDMats.include

PDMats.invquad — Function.

invquad(a::AbstractPDMat, x::StridedVecOrMat)

Return the value of the quadratic form defined by inv(a) applied to x.

For most PDMat types this is done in a way that does not require evaluation of inv(a).

If x is a vector the quadratic form is x' * a * x. If x is a matrix the quadratic form is applied column-wise.

PDMats.invquad! — Function.

invquad!(r::AbstractArray, a::AbstractPDMat, x::StridedMatrix)

Overwrite r with the value of the quadratic form defined by inv(a) applied columnwise to x

PDMats.pdadd

PDMats.pdadd!

PDMats.pdtest_add

PDMats.pdtest_basics

PDMats.pdtest_cmat

PDMats.pdtest_diag

PDMats.pdtest_eig

PDMats.pdtest_logdet

PDMats.pdtest_mul

PDMats.pdtest_quad

PDMats.pdtest_rdiv

PDMats.pdtest_scale

PDMats.pdtest_triprod

PDMats.pdtest_whiten

PDMats.quad — Function.

quad(a::AbstractPDMat, x::StridedVecOrMat)

Return the value of the quadratic form defined by a applied to x

If x is a vector the quadratic form is x' * a * x. If x is a matrix the quadratic form is applied column-wise.

PDMats.quad! — Function.

quad!(r::AbstractArray, a::AbstractPDMat, x::StridedMatrix)

Overwrite r with the value of the quadratic form defined by a applied columnwise to x

PDMats.test_pdmat

PDMats.unwhiten

PDMats.unwhiten!

PDMats.whiten — Function.

whiten(a::AbstractPDMat, x::StridedVecOrMat)
whiten!(a::AbstractPDMat, x::StridedVecOrMat)
whiten!(r::StridedVecOrMat, a::AbstractPDMat, x::StridedVecOrMat)
unwhiten(a::AbstractPDMat, x::StridedVecOrMat)
unwhiten!(a::AbstractPDMat, x::StridedVecOrMat)
unwhiten!(r::StridedVecOrMat, a::AbstractPDMat, x::StridedVecOrMat)

Allocating and in-place versions of the whitening transform (or its inverse) defined by a applied to x

If the covariance of x is a then the covariance of the result will be I. The name whiten indicates that this function transforms a correlated multivariate random variable to "white noise".

julia> using PDMats

julia> X = vcat(ones(4)', (1:4)')
2×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  2.0  3.0  4.0

julia> a = PDMat(X * X')
PDMat{Float64,Array{Float64,2}}(2, [4.0 10.0; 10.0 30.0], Cholesky{Float64,Array{Float64,2}}([2.0 5.0; 10.0 2.23607], 'U', 0))

julia> W = whiten(a, X)
2×4 Array{Float64,2}:
  0.5       0.5       0.5       0.5    
 -0.67082  -0.223607  0.223607  0.67082

julia> W * W'
2×2 Array{Float64,2}:
 1.0  0.0
 0.0  1.0

PDMats.whiten!

PDMats.wsumsq