Docstrings

OnlineStatsBase.Bounded — Type.

Bounded(w::Weight, λ::Float64)

Bound the weight by a constant.

$γ′(t) = max(γ(t), λ)$

OnlineStatsBase.EqualWeight — Type.

EqualWeight()

Equally weighted observations.

$γ(t) = 1 / t$

OnlineStatsBase.ExponentialWeight — Type.

ExponentialWeight(λ::Float64)
ExponentialWeight(lookback::Int)

Exponentially weighted observations. The first weight is 1.0 and all else are λ = 2 / (lookback + 1).

$γ(t) = λ$

OnlineStatsBase.HarmonicWeight — Type.

HarmonicWeight(a = 10.0)

Weight determined by harmonic series.

$γ(t) = a / (a + t - 1)$

OnlineStatsBase.LearningRate — Type.

LearningRate(r = .6)

Slowly decreasing weight. Satisfies the standard stochastic approximation assumption $∑ γ(t) = ∞, ∑ γ(t)^2 < ∞$ if $r ∈ (.5, 1]$.

$γ(t) = inv(t ^ r)$

OnlineStatsBase.LearningRate2 — Type.

LearningRate2(c = .5)

Slowly decreasing weight.

$γ(t) = inv(1 + c * (t - 1))$

OnlineStatsBase.McclainWeight — Type.

McclainWeight(α = .1)

Weight which decreases into a constant.

$γ(t) = γ(t-1) / (1 + γ(t-1) - α)$

OnlineStatsBase.OnlineIterator

OnlineStatsBase.OnlineStat

OnlineStatsBase.OnlineStatsBase

OnlineStatsBase.Scaled — Type.

Scaled(w::Weight, λ::Float64)

Scale a weight by a constant.

$γ′(t) = λ * γ(t)$

OnlineStatsBase.Weight

OnlineStatsBase.XY

OnlineStatsBase._fit!

OnlineStatsBase._merge!

OnlineStatsBase.change_denom

OnlineStatsBase.copycol!

OnlineStatsBase.copyrow!

OnlineStatsBase.eachcol — Function.

eachcol(x::AbstractMatrix)
eachcol(x::AbstractMatrix, y::AbstractVector)

Iterator over the columns of x (paired in a tuple with the values of y).

Example

for xi in eachcol(rand(3,2))
    println(xi)
end

OnlineStatsBase.eachrow — Function.

eachrow(x::AbstractMatrix)
eachrow(x::AbstractMatrix, y::AbstractVector)

Iterator over the rows of x (paired in a tuple with the values of y).

Example

for xi in eachrow(rand(3,2))
    println(xi)
end

OnlineStatsBase.eval

StatsBase.fit! — Function.

fit!(stat::OnlineStat, data)

Update the "sufficient statistics" of a stat with more data. If typeof(data) is not the type of a single observation for the provided stat, fit! will attempt to iterate through and fit! each item in data. Therefore, fit!(Mean(), 1:10) translates roughly to:

o = Mean()
for x in 1:10
    fit!(o, x)
end

OnlineStatsBase.include

OnlineStatsBase.name

OnlineStatsBase.nobs

OnlineStatsBase.smooth — Function.

smooth(a, b, γ)

Weighted average of a and b with weight γ.

OnlineStatsBase.smooth! — Function.

smooth!(a, b, γ)

Update a in place by applying the smooth function elementwise with b.

OnlineStatsBase.smooth_syr! — Function.

smooth_syr!(A::AbstractMatrix, x, γ::Number)

Weighted average of symmetric rank-1 update. Updates the upper triangle of:

A = (1 - γ) * A + γ * x * x'

LearnBase.value — Function.

value(o::OnlineStat)

Calculate the value of the stat from its "sufficient statistics".