Docstrings

StatPlots.@animate

StatPlots.@colorant_str

StatPlots.@df — Macro.

`@df d x`

Convert every symbol in the expression x with the respective column in d if it exists.

If you want to avoid replacing the symbol, escape it with ^.

NA values are replaced with NaN for columns of Float64 and "" or Symbol() for strings and symbols respectively.

x can be either a plot command or a block of plot commands.

`@df x`

Curried version of @df d x. Outputs an anonymous function d -> @df d x.

StatPlots.@gif

StatPlots.@layout

StatPlots.@recipe

StatPlots.@series

StatPlots.@shorthands

StatPlots.@userplot

StatPlots.ABGR

StatPlots.ADIN99

StatPlots.ADIN99d

StatPlots.ADIN99o

StatPlots.AGray

StatPlots.AGray32

StatPlots.AHSI

StatPlots.AHSL

StatPlots.AHSV

StatPlots.ALCHab

StatPlots.ALCHuv

StatPlots.ALMS

StatPlots.ALab

StatPlots.ALuv

StatPlots.ARGB

StatPlots.ARGB32

StatPlots.AXYZ

StatPlots.AYCbCr

StatPlots.AYIQ

StatPlots.AbstractBackend

StatPlots.AbstractGray

StatPlots.AbstractLayout

StatPlots.AbstractPlot

StatPlots.AbstractRGB

StatPlots.AlphaColor

StatPlots.AndrewsPlot

StatPlots.Animation

StatPlots.AxyY

StatPlots.BGR

StatPlots.BGRA

StatPlots.BezierCurve

StatPlots.CIE1931JV_CMF

StatPlots.CIE1931J_CMF

StatPlots.CIE1931_CMF

StatPlots.CIE1964_CMF

StatPlots.Color

StatPlots.Color3

StatPlots.ColorAlpha

StatPlots.ColorGradient

StatPlots.ColorTypes

StatPlots.Colorant

StatPlots.ColorantNormed

StatPlots.Colors

StatPlots.CornerPlot

StatPlots.CorrPlot

StatPlots.DE_2000

StatPlots.DE_94

StatPlots.DE_AB

StatPlots.DE_BFD

StatPlots.DE_CMC

StatPlots.DE_DIN99

StatPlots.DE_DIN99d

StatPlots.DE_DIN99o

StatPlots.DE_JPC79

StatPlots.DIN99

StatPlots.DIN99A

StatPlots.DIN99d

StatPlots.DIN99dA

StatPlots.DIN99o

StatPlots.DIN99oA

StatPlots.Formatted

StatPlots.Fractional

StatPlots.GR

StatPlots.Gray

StatPlots.Gray24

StatPlots.GrayA

StatPlots.GroupedBar

StatPlots.HSB

StatPlots.HSI

StatPlots.HSIA

StatPlots.HSL

StatPlots.HSLA

StatPlots.HSV

StatPlots.HSVA

StatPlots.KW

StatPlots.LCHab

StatPlots.LCHabA

StatPlots.LCHuv

StatPlots.LCHuvA

StatPlots.LMS

StatPlots.LMSA

StatPlots.Lab

StatPlots.LabA

StatPlots.Luv

StatPlots.LuvA

StatPlots.MSC

StatPlots.OHLC

StatPlots.PlotThemes

StatPlots.PlotUtils

StatPlots.Plots

StatPlots.QQNorm

StatPlots.QQPlot

StatPlots.RGB

StatPlots.RGB1

StatPlots.RGB24

StatPlots.RGB4

StatPlots.RGBA

StatPlots.RecipeData

StatPlots.RecipesBase

StatPlots.Segments

StatPlots.Shape

StatPlots.StatPlots

StatPlots.Surface

StatPlots.Transparent3

StatPlots.TransparentColor

StatPlots.TransparentGray

StatPlots.TransparentRGB

StatPlots.XYZ

StatPlots.XYZA

StatPlots.YCbCr

StatPlots.YCbCrA

StatPlots.YIQ

StatPlots.YIQA

StatPlots.__init__

StatPlots._arg2string

StatPlots._argnames

StatPlots._violin_warned

StatPlots.adapted_grid

StatPlots.add_backend

StatPlots.add_label

StatPlots.add_theme

StatPlots.aliases

StatPlots.alpha

StatPlots.alphacolor

StatPlots.andrewsplot — Function.

andrewsplot(args...; kw...)

Shows each row of an array (or table) as a line. The x argument specifies a grouping variable. This is a way to visualize structure in high-dimensional data. https://en.wikipedia.org/wiki/Andrews_plot #Examples

using RDatasets, StatPlots
iris = dataset("datasets", "iris")
@df iris andrewsplot(:Species, cols(1:4))

StatPlots.andrewsplot!

StatPlots.animate

StatPlots.annotate!

StatPlots.arrow

StatPlots.attr!

StatPlots.backend

StatPlots.backend_name

StatPlots.backend_object

StatPlots.backends

StatPlots.bar

StatPlots.bar!

StatPlots.barh

StatPlots.barh!

StatPlots.barhist

StatPlots.barhist!

StatPlots.base_color_type

StatPlots.base_colorant_type

StatPlots.bbox

StatPlots.bkfe — Function.

binned kernel function estimator

StatPlots.blue

StatPlots.boxplot

StatPlots.boxplot!

StatPlots.brush

StatPlots.ccolor

StatPlots.cdensity

StatPlots.cdensity!

StatPlots.center

StatPlots.cgrad

StatPlots.cgradients

StatPlots.cie_color_match

StatPlots.clibraries

StatPlots.clibrary

StatPlots.closeall

StatPlots.color

StatPlots.color_type

StatPlots.coloralpha

StatPlots.colordiff

StatPlots.colormap

StatPlots.colormatch

StatPlots.column_names

StatPlots.column_types

StatPlots.comp1

StatPlots.comp2

StatPlots.comp3

StatPlots.compute_name

StatPlots.contour

StatPlots.contour!

StatPlots.contour3d

StatPlots.contour3d!

StatPlots.contourf

StatPlots.contourf!

StatPlots.convert_missing

StatPlots.coords

StatPlots.cornerplot

StatPlots.cornerplot!

StatPlots.corrplot

StatPlots.corrplot!

StatPlots.current

StatPlots.curve_points

StatPlots.curves

StatPlots.curves!

StatPlots.cvec

StatPlots.dataviewer

StatPlots.default

StatPlots.default_cgrad

StatPlots.default_range

StatPlots.density

StatPlots.density!

StatPlots.deuteranopic

StatPlots.df_helper

StatPlots.distinguishable_colors

StatPlots.diverging_palette

StatPlots.ea_binnumber

StatPlots.ea_histogram

StatPlots.ea_histogram!

StatPlots.eval

StatPlots.extract_columns_from_iterabletable

StatPlots.font

StatPlots.frame

StatPlots.gamutmax

StatPlots.gamutmin

StatPlots.get_col_from_dict

StatPlots.get_color_palette

StatPlots.gif

StatPlots.gr

StatPlots.gray

StatPlots.green

StatPlots.grid

StatPlots.grouped_xy

StatPlots.groupedbar

StatPlots.groupedbar!

StatPlots.gui

StatPlots.hdf5

StatPlots.heatmap

StatPlots.heatmap!

StatPlots.hex

StatPlots.hexbin

StatPlots.hexbin!

StatPlots.histogram

StatPlots.histogram!

StatPlots.histogram2d

StatPlots.histogram2d!

StatPlots.hline

StatPlots.hline!

StatPlots.hspan

StatPlots.hspan!

StatPlots.include

StatPlots.inline

StatPlots.insert_kw!

StatPlots.inspectdr

StatPlots.invisible

StatPlots.isdark

StatPlots.iter_segments

StatPlots.linbin

StatPlots.loc

StatPlots.mapc

StatPlots.mapreducec

StatPlots.marginalhist

StatPlots.marginalhist!

StatPlots.mov

StatPlots.mp4

StatPlots.not_kw

StatPlots.notch_width

StatPlots.ohlc

StatPlots.ohlc!

StatPlots.optimize_datetime_ticks

StatPlots.optimize_ticks

StatPlots.palette

StatPlots.parse_iterabletable_call!

StatPlots.path3d

StatPlots.path3d!

StatPlots.pgfplots

StatPlots.pie

StatPlots.pie!

StatPlots.plot

StatPlots.plot!

StatPlots.plot3d

StatPlots.plot3d!

StatPlots.plot_color

StatPlots.plotarea

StatPlots.plotattr

StatPlots.plotly

StatPlots.plotlyjs

StatPlots.plots_heatmap

StatPlots.plots_heatmap!

StatPlots.png

StatPlots.portfoliocomposition

StatPlots.portfoliocomposition!

StatPlots.protanopic

StatPlots.pyplot

StatPlots.qqnorm

StatPlots.qqnorm!

StatPlots.qqplot

StatPlots.qqplot!

StatPlots.quiver

StatPlots.quiver!

StatPlots.red

StatPlots.reducec

StatPlots.rgb_string

StatPlots.rgba_string

StatPlots.rotate

StatPlots.rotate!

StatPlots.savefig

StatPlots.scatter

StatPlots.scatter!

StatPlots.scatter3d

StatPlots.scatter3d!

StatPlots.scatterhist

StatPlots.scatterhist!

StatPlots.sequential_palette

StatPlots.set_theme

StatPlots.shape_coords

StatPlots.showgradient

StatPlots.showgradient!

StatPlots.showlibrary

StatPlots.showlibrary!

StatPlots.showtheme

StatPlots.showtheme!

StatPlots.spy

StatPlots.spy!

StatPlots.stephist

StatPlots.stephist!

StatPlots.sticks

StatPlots.sticks!

StatPlots.stringify

StatPlots.stroke

StatPlots.surface

StatPlots.surface!

StatPlots.test_examples

StatPlots.testhist

StatPlots.testhist!

StatPlots.text

StatPlots.theme

StatPlots.title!

StatPlots.translate

StatPlots.translate!

StatPlots.treepositions

StatPlots.tritanopic

StatPlots.twinx

StatPlots.unicodeplots

StatPlots.update_ticks_guides

StatPlots.violin

StatPlots.violin!

StatPlots.violin_coords

StatPlots.vline

StatPlots.vline!

StatPlots.vspan

StatPlots.vspan!

StatPlots.wand_bins — Function.

Returns optimal histogram bin widths in accordance to Wand (1995)'s criterion'

StatPlots.weighted_color_mean

StatPlots.whitebalance

StatPlots.wireframe

StatPlots.wireframe!

StatPlots.with

StatPlots.wrap

StatPlots.xaxis!

StatPlots.xflip!

StatPlots.xgrid!

StatPlots.xlabel!

StatPlots.xlims

StatPlots.xlims!

StatPlots.xticks!

StatPlots.xyY

StatPlots.xyYA

StatPlots.yaxis!

StatPlots.yflip!

StatPlots.ygrid!

StatPlots.ylabel!

StatPlots.ylims

StatPlots.ylims!

StatPlots.yticks!

StatPlots.yz_args

StatPlots.zlims

StatPlots.zlims!