Docstrings

GLPK.@check

GLPK.@check!

GLPK.@glpk_ccall

GLPK.ASN_MAX

GLPK.ASN_MIN

GLPK.ASN_MMP

GLPK.AbstractCallbackData — Type.

AbstractCallbackData

An abstract type to prevent recursive type definition of Optimizer and CallbackData, each of which need the other type in a field.

GLPK.Attr

GLPK.BF_BG

GLPK.BF_BTF

GLPK.BF_FT

GLPK.BF_GR

GLPK.BF_LUF

GLPK.BR_DTH

GLPK.BR_FFV

GLPK.BR_LFV

GLPK.BR_MFV

GLPK.BR_PCH

GLPK.BS

GLPK.BT_BFS

GLPK.BT_BLB

GLPK.BT_BPH

GLPK.BT_DFS

GLPK.BV

GLPK.BasisFactParam

GLPK.CV

GLPK.CallbackData — Type.

CallbackData

GLPK.CallbackFunction — Type.

CallbackFunction

The attribute to set the callback function in GLPK. The function takes a single argument of type CallbackData.

GLPK.DB

GLPK.DBL_MAX

GLPK.DN_BRNCH

GLPK.DUAL

GLPK.DUALP

GLPK.EBADB

GLPK.EBOUND

GLPK.ECOND

GLPK.EDATA

GLPK.EFAIL

GLPK.EINSTAB

GLPK.EITLIM

GLPK.EMIPGAP

GLPK.ENOCVG

GLPK.ENODFS

GLPK.ENOFEAS

GLPK.ENOPFS

GLPK.EOBJLL

GLPK.EOBJUL

GLPK.ERANGE

GLPK.EROOT

GLPK.ESING

GLPK.ESTOP

GLPK.ETMLIM

GLPK.FEAS

GLPK.FR

GLPK.FX

GLPK.GLPK

GLPK.GLPKError

GLPK.GLPKFatalError

GLPK.IBINGO

GLPK.IBRANCH

GLPK.ICUTGEN

GLPK.IHEUR

GLPK.INFEAS

GLPK.IPREPRO

GLPK.IPT

GLPK.IROWGEN

GLPK.ISELECT

GLPK.IV

GLPK.InteriorParam

GLPK.IntoptParam

GLPK.KKT_CS

GLPK.KKT_DB

GLPK.KKT_DE

GLPK.KKT_PB

GLPK.KKT_PE

GLPK.LO

GLPK.LQOI

GLPK.MAX

GLPK.MIN

GLPK.MIP

GLPK.MOI

GLPK.MPS_DECK

GLPK.MPS_FILE

GLPK.MSG_ALL

GLPK.MSG_DBG

GLPK.MSG_ERR

GLPK.MSG_OFF

GLPK.MSG_ON

GLPK.MathProgWorkspace

GLPK.NF

GLPK.NL

GLPK.NOFEAS

GLPK.NO_BRNCH

GLPK.NS

GLPK.NU

GLPK.OFF

GLPK.ON

GLPK.OPT

GLPK.ORD_AMD

GLPK.ORD_NONE

GLPK.ORD_QMD

GLPK.ORD_SYMAMD

GLPK.Optimizer

GLPK.PP_ALL

GLPK.PP_NONE

GLPK.PP_ROOT

GLPK.PRIMAL

GLPK.PT_PSE

GLPK.PT_STD

GLPK.Param

GLPK.Prob

GLPK.RF_CLQ

GLPK.RF_COV

GLPK.RF_CUT

GLPK.RF_GMI

GLPK.RF_LAZY

GLPK.RF_MIR

GLPK.RF_REG

GLPK.ROW_SENSE_MAP

GLPK.RT_HAR

GLPK.RT_STD

GLPK.SF_2N

GLPK.SF_AUTO

GLPK.SF_EQ

GLPK.SF_GM

GLPK.SF_SKIP

GLPK.SOL

GLPK.SUPPORTED_CONSTRAINTS

GLPK.SUPPORTED_OBJECTIVES

GLPK.SimplexParam

GLPK.UNBND

GLPK.UNDEF

GLPK.UP

GLPK.UP_BRNCH

GLPK.VARIABLE_TYPE_MAP

GLPK.VecOrNothing

GLPK.__init__

GLPK.__internal_callback__ — Function.

__internal_callback__(tree::Ptr{Cvoid}, info::Ptr{Cvoid})

Dummy callback function for internal use only. Responsible for updating the objective bound.

GLPK._add_obj

GLPK._adv_basis_flags

GLPK._alloc_size

GLPK._bf_exists

GLPK._bounds_are_valid

GLPK._bounds_type_is_valid

GLPK._can_branch

GLPK._certificates_potentially_available — Function.

_certificates_potentially_available(model::Optimizer)

Return true if an infeasiblity certificate or an unbounded ray is potentially available (i.e., the model has been solved using either the Simplex or Exact methods).

GLPK._col_is_valid

GLPK._col_is_valid_w0

GLPK._cols_ids_content

GLPK._cols_ids_size

GLPK._constr_type_is_valid

GLPK._copy_names_flag

GLPK._del_all_objs

GLPK._del_obj

GLPK._dir_is_valid

GLPK._eps_is_valid

GLPK._err_hook

GLPK._file_is_readable

GLPK._file_is_writable

GLPK._indices_vectors_dup

GLPK._init_env_succeeded

GLPK._ios_add_row_flags

GLPK._ios_node_is_active

GLPK._ios_node_is_valid

GLPK._ios_row_is_valid

GLPK._is_dual_feasible

GLPK._is_prim_feasible

GLPK._kind_is_valid

GLPK._kkt_cond_param

GLPK._klass_is_valid

GLPK._list_ids

GLPK._lp_param

GLPK._mpl_workspace

GLPK._mps_format

GLPK._mps_param

GLPK._obj_dir_is_valid

GLPK._open_tee_succeeded

GLPK._pointer_is_valid

GLPK._print_ranges_flags

GLPK._prob

GLPK._read_prob_flags

GLPK._reason

GLPK._row_is_valid

GLPK._rowcol_is_valid

GLPK._rows_and_cols

GLPK._rows_ids_content

GLPK._rows_ids_size

GLPK._scale_flags

GLPK._sel_is_valid

GLPK._sol_param

GLPK._stat_is_valid

GLPK._status_is_optimal

GLPK._string_length

GLPK._succeeded

GLPK._term_out_flag

GLPK._throw_invalid_method — Function.

_throw_invalid_method(instance::Optimizer)

A helper function to throw an error when the method is set incorrectly. Mainly used to enforce type-stability in functions that have a run-time switch on the method.

GLPK._tree

GLPK._var_is_basic

GLPK._var_is_non_basic

GLPK._vectors_all_same_size

GLPK._vectors_size

GLPK._write_prob_flags

GLPK.add_cols

GLPK.add_lazy_constraint! — Function.

add_lazy_constraint!(cb_data::GLPK.CallbackData, func::LQOI.Linear, set::S) where S <: Union{LQOI.LE, LQOI.GE, LQOI.EQ}

Add a lazy constraint to the model cb_data.model. This can only be called in a callback from GLPK.IROWGEN.

GLPK.add_row! — Function.

add_row!(problem::GLPK.Prob, columns::Vector{Int},
         coefficients::Vector{Float64}, sense::Cchar, rhs::Real)

Helper function to add a row to the problem. Sense must be one of 'E' (ax == b), 'G' (ax >= b), 'L' (ax <= b) , or 'R' (b <= ax).

If the sense is 'R' the rhs should be the lower bound, and the bounds should be set in a new API call to enforce the upper bound.

GLPK.add_rows

GLPK.adv_basis

GLPK.analyze_bound

GLPK.analyze_coef

GLPK.bf_exists

GLPK.bf_updated

GLPK.btran

GLPK.calloc

GLPK.cdouble_vec

GLPK.change_row_sense! — Function.

change_row_sense!(model::Optimizer, row, sense)

Convert a linear constraint into another type of linear constraint by changing the comparison sense.

Constraint types supported are 'E' (equality), 'L' (less-than), and 'G' (greater-than).

For example, ax <= b can become ax >= b or ax == b.

GLPK.check_cnfsat

GLPK.check_deps

GLPK.check_dup

GLPK.check_glpk_version

GLPK.check_kkt

GLPK.cint_vec

GLPK.close_tee

GLPK.copy_function_result! — Function.

copy_function_result!(dest::Vector, foo, model::GLPK.Prob)

A helper function that loops through the indices in dest and stores the result of foo(model, i) for the ith index.

GLPK.copy_prob

GLPK.cpx_basis

GLPK.create_index

GLPK.del_cols

GLPK.del_rows

GLPK.delete_index

GLPK.delete_prob

GLPK.dual_rtest

GLPK.erase_prob

GLPK.eval

GLPK.eval_tab_col

GLPK.eval_tab_row

GLPK.exact

GLPK.factorize

GLPK.find_col

GLPK.find_row

GLPK.free

GLPK.free_env

GLPK.ftran

GLPK.get_bfcp

GLPK.get_bhead

GLPK.get_col_bind

GLPK.get_col_bound_type — Function.

get_col_bound_type(lower::Float64, upper::Float64)

Return the GLPK type of the variable bound given a lower bound of lower and an upper bound of upper.

GLPK.get_col_dual

GLPK.get_col_kind

GLPK.get_col_lb

GLPK.get_col_name

GLPK.get_col_prim

GLPK.get_col_stat

GLPK.get_col_type

GLPK.get_col_ub

GLPK.get_dual_stat

GLPK.get_infeasibility_ray — Function.

get_infeasibility_ray(model::Optimizer, ray::Vector{Float64})

Get the Farkas certificate of primal infeasiblity.

Can only be called when GLPK.simplex is used as the solver.

GLPK.get_mat_col

GLPK.get_mat_row

GLPK.get_num_bin

GLPK.get_num_cols

GLPK.get_num_int

GLPK.get_num_nz

GLPK.get_num_rows

GLPK.get_obj_coef

GLPK.get_obj_dir

GLPK.get_obj_name

GLPK.get_obj_val

GLPK.get_prim_stat

GLPK.get_prob_name

GLPK.get_rii

GLPK.get_row_bind

GLPK.get_row_dual

GLPK.get_row_lb

GLPK.get_row_name

GLPK.get_row_prim

GLPK.get_row_stat

GLPK.get_row_type

GLPK.get_row_ub

GLPK.get_sjj

GLPK.get_status — Function.

get_status(model::Optimizer)

Get the status from GLPK depending on which method was used to solve the model.

GLPK.get_unbnd_ray

GLPK.get_unbounded_ray — Function.

get_unbounded_ray(model::Optimizer, ray::Vector{Float64})

Get the certificate of primal unboundedness.

Can only be called when GLPK.simplex is used as the solver.

GLPK.get_variable_types — Function.

get_variable_types(model::Optimizer)

Return a vector of symbols (one element for each variable) of the variable type. The symbols are given by the key-value mapping in GLPK.VARIABLE_TYPE_MAP.

GLPK.glpkdefver

GLPK.glpkmaxver

GLPK.glpkminver

GLPK.glpkwinver

GLPK.include

GLPK.init_env

GLPK.init_iocp

GLPK.init_iptcp

GLPK.init_smcp

GLPK.interior

GLPK.intfeas1

GLPK.intopt

GLPK.ios_add_row

GLPK.ios_best_node

GLPK.ios_branch_upon

GLPK.ios_can_branch

GLPK.ios_clear_pool

GLPK.ios_curr_node

GLPK.ios_del_row

GLPK.ios_get_prob

GLPK.ios_heur_sol

GLPK.ios_mip_gap

GLPK.ios_next_node

GLPK.ios_node_bound

GLPK.ios_node_data

GLPK.ios_node_level

GLPK.ios_pool_size

GLPK.ios_prev_node

GLPK.ios_reason

GLPK.ios_row_attr

GLPK.ios_select_node

GLPK.ios_terminate

GLPK.ios_tree_size

GLPK.ios_up_node

GLPK.ipt_col_dual

GLPK.ipt_col_prim

GLPK.ipt_obj_val

GLPK.ipt_row_dual

GLPK.ipt_row_prim

GLPK.ipt_status

GLPK.jl_get_preemptive_check

GLPK.jl_obj_is_valid

GLPK.jl_set_preemptive_check

GLPK.libglpk

GLPK.load_matrix

GLPK.load_variable_primal! — Function.

load_variable_primal!(cb_data::CallbackData)

Load the variable primal solution in a callback.

This can only be called in a callback from GLPK.IROWGEN. After it is called, you can access the VariablePrimal attribute as usual.

GLPK.malloc

GLPK.mem_limit

GLPK.mem_usage

GLPK.minisat1

GLPK.mip_col_val

GLPK.mip_obj_val

GLPK.mip_row_val

GLPK.mip_status

GLPK.mpl_build_prob

GLPK.mpl_free_wksp

GLPK.mpl_generate

GLPK.mpl_postsolve

GLPK.mpl_read_data

GLPK.mpl_read_model

GLPK.open_tee

GLPK.prim_rtest

GLPK.print_ipt

GLPK.print_mip

GLPK.print_ranges

GLPK.print_sol

GLPK.read_cnfsat

GLPK.read_ipt

GLPK.read_lp

GLPK.read_mip

GLPK.read_mps

GLPK.read_prob

GLPK.read_sol

GLPK.round_bounds_to_integer — Function.

round_bounds_to_integer(model)::Tuple{Bool, Vector{Float64}, Vector{Float64}}

GLPK does not allow integer variables with fractional bounds. Therefore, we round the bounds of binary and integer variables to integer values prior to solving.

Returns a tuple of the original bounds, along with a Boolean flag indicating if they need to be reset after solve.

GLPK.scale_prob

GLPK.set_bfcp

GLPK.set_col_bnds

GLPK.set_col_kind

GLPK.set_col_name

GLPK.set_col_stat

GLPK.set_mat_col

GLPK.set_mat_row

GLPK.set_obj_coef

GLPK.set_obj_dir

GLPK.set_obj_name

GLPK.set_parameter — Function.

set_parameter(param_store, key::Symbol, value)

Set the field name key in a param_store type (that is one of InteriorParam, IntoptParam, or SimplexParam) to value.

GLPK.set_prob_name

GLPK.set_rii

GLPK.set_row_bnds

GLPK.set_row_name

GLPK.set_row_stat

GLPK.set_sjj

GLPK.set_variable_bound — Function.

set_variable_bound(model::Optimizer, column::Int, lower::Float64, upper::Float64)

Set the bounds of the variable in column column to [lower, upper].

GLPK.simplex

GLPK.sort_matrix

GLPK.std_basis

GLPK.term_out

GLPK.transform_col

GLPK.transform_row

GLPK.unscale_prob

GLPK.vecornothing_length

GLPK.version

GLPK.warm_up

GLPK.write_cnfsat

GLPK.write_ipt

GLPK.write_lp

GLPK.write_mip

GLPK.write_mps

GLPK.write_prob

GLPK.write_sol