Docstrings

ProgressMeter.@showprogress — Macro.

@showprogress dt "Computing..." for i = 1:50
    # computation goes here
end

@showprogress dt "Computing..." pmap(x->x^2, 1:50)

displays progress in performing a computation. dt is the minimum interval between updates to the user. You may optionally supply a custom message to be printed that specifies the computation being performed.

@showprogress works for loops, comprehensions, map, and pmap.

ProgressMeter.AbstractProgress

ProgressMeter.BarGlyphs — Type.

Holds the five characters that will be used to generate the progress bar.

ProgressMeter.Progress — Type.

prog = Progress(n; dt=0.1, desc="Progress: ", color=:green, output=stderr, barlen=tty_width(desc)) creates a progress meter for a task with n iterations or stages. Output will be generated at intervals at least dt seconds apart, and perhaps longer if each iteration takes longer than dt. desc is a description of the current task.

ProgressMeter.ProgressMeter — Module.

ProgressMeter contains a suite of utilities for displaying progress in long-running computations. The major functions/types in this module are:

@showprogress: an easy interface for straightforward situations

Progress: an object for managing progress updates with a predictable number of iterations

ProgressThresh: an object for managing progress updates where termination is governed by a threshold

next! and update!: report that progress has been made

cancel and finish!: early termination

ProgressMeter.ProgressThresh — Type.

prog = ProgressThresh(thresh; dt=0.1, desc="Progress: ", color=:green, output=stderr) creates a progress meter for a task which will terminate once a value less than or equal to thresh is reached. Output will be generated at intervals at least dt seconds apart, and perhaps longer if each iteration takes longer than dt. desc is a description of the current task.

ProgressMeter.ProgressUnknown — Type.

prog = ProgressUnknown(; dt=0.1, desc="Progress: ", color=:green, output=stderr) creates a progress meter for a task which has a non-deterministic termination criterion. Output will be generated at intervals at least dt seconds apart, and perhaps longer if each iteration takes longer than dt. desc is a description of the current task.

ProgressMeter.ProgressWrapper

ProgressMeter.barstring

ProgressMeter.cancel — Function.

cancel(prog, [msg], [color=:red]) cancels the progress display before all tasks were completed. Optionally you can specify the message printed and its color.

See also finish!.

ProgressMeter.compute_front

ProgressMeter.durationstring

ProgressMeter.eval

ProgressMeter.finish! — Function.

finish!(prog) indicates that all tasks have been completed.

See also cancel.

ProgressMeter.include

ProgressMeter.move_cursor_up_while_clearing_lines

ProgressMeter.ncalls — Function.

Infer the number of calls to the mapped function (i.e. the length of the returned array) given the input arguments to map or pmap.

ProgressMeter.next! — Function.

next!(prog, [color]) reports that one unit of progress has been made. Depending on the time interval since the last update, this may or may not result in a change to the display.

You may optionally change the color of the display. See also update!.

ProgressMeter.printover

ProgressMeter.printvalues!

ProgressMeter.progress_map — Function.

progress_map(f, c...; mapfun=map, progress=Progress(...), kwargs...)

Run a map-like function while displaying progress.

mapfun can be any function, but it is only tested with map and pmap.

ProgressMeter.progress_pmap — Function.

progress_pmap(f, [::AbstractWorkerPool], c...; progress=Progress(...), kwargs...)

Run pmap while displaying progress.

ProgressMeter.showprogress_process_expr

ProgressMeter.tty_width

ProgressMeter.update! — Function.

update!(prog, counter, [color]) sets the progress counter to counter, relative to the n units of progress specified when prog was initialized. Depending on the time interval since the last update, this may or may not result in a change to the display.

If prog is a ProgressThresh, update!(prog, val, [color]) specifies the current value.

You may optionally change the color of the display. See also next!.

ProgressMeter.updateProgress!