Docstrings

TimeZones.@optional — Macro.

@optional(expr)

Creates multiple method signatures to allow optional arguments before required arguments. For example:

f(a=1, b=2, c) = ...

becomes:

f(a, b, c) = ...
f(a, c) = f(a, 2, c)
f(c) = f(1, 2, c)

TimeZones.@tz_str — Macro.

@tz_str -> TimeZone

Constructs a TimeZone subtype based upon the string at parse time. See docstring of TimeZone for more details.

julia> tz"Africa/Nairobi"
Africa/Nairobi (UTC+3)

TimeZones.ARCHIVE_DIR

TimeZones.AmbiguousTimeError — Type.

AmbiguousTimeError(local_datetime, timezone)

The provided local datetime is ambiguous within the specified timezone. Typically occurs on daylight saving time transitions which "fall back" causing duplicate hour long period.

TimeZones.COMPILED_DIR

TimeZones.DEPS_DIR

Dates.DateTime — Type.

DateTime(::ZonedDateTime) -> DateTime

Returns an equivalent DateTime without any TimeZone information.

TimeZones.FIXED_TIME_ZONE_REGEX

TimeZones.FixedTimeZone — Type.

FixedTimeZone

A TimeZone with a constant offset for all of time.

TimeZones.ISOZonedDateTimeFormat

TimeZones.Local

TimeZones.NonExistentTimeError — Type.

NonExistentTimeError(local_datetime, timezone)

The provided local datetime is does not exist within the specified timezone. Typically occurs on daylight saving time transitions which "spring forward" causing an hour long period to be skipped.

TimeZones.PKG_DIR

TimeZones.TIME_ZONES

TimeZones.TZData

TimeZones.TZFILE_MAX

TimeZones.TZ_REGEX

TimeZones.TZ_SOURCE_DIR

TimeZones.TimeError

Dates.TimeZone — Type.

TimeZone(str::AbstractString) -> TimeZone

Constructs a TimeZone subtype based upon the string. If the string is a recognized time zone name then data is loaded from the compiled IANA time zone database. Otherwise the string is assumed to be a static time zone.

A list of recognized time zones names is available from timezone_names(). Supported static time zone string formats can be found in FixedTimeZone(::AbstractString).

TimeZones.TimeZones

TimeZones.Transition

TimeZones.TransitionTimeInfo

TimeZones.UTCOffset — Type.

UTCOffset

A UTCOffset is an amount of time subtracted from or added to UTC to get the current local time – whether it's standard time or daylight saving time.

TimeZones.UnhandledTimeError — Type.

UnhandledTimeError(timezone)

The timezone calculation occurs beyond the last calculated transition.

TimeZones.VariableTimeZone — Type.

VariableTimeZone

A TimeZone with an offset that changes over time.

TimeZones.ZonedDateTime — Type.

ZonedDateTime(y, [m, d, h, mi, s, ms], tz, [amb]) -> DateTime

Construct a ZonedDateTime type by parts. Arguments y, m, ..., ms must be convertible to Int64 and tz must be a TimeZone. If the given DateTime is ambiguous in the given TimeZone then amb can be supplied to resolve ambiguity.

TimeZones.__init__

TimeZones.abbreviation

TimeZones.all_timezones — Function.

all_timezones() -> Array{TimeZone}

Returns all pre-computed TimeZones.

all_timezones(criteria::Function) -> Array{TimeZone}

Returns TimeZones that match the given criteria function. The criteria function takes two parameters: UTC transition (DateTime) and transition zone (FixedTimeZone).

Examples

Find all time zones which contain an absolute UTC offset greater than 15 hours:

all_timezones() do dt, zone
    abs(zone.offset.std) > Dates.Second(Dates.Hour(15))
end

Determine all time zones which have a non-hourly daylight saving time offset:

all_timezones() do dt, zone
    zone.offset.dst % Dates.Second(Dates.Hour(1)) != 0
end

TimeZones.astimezone — Function.

astimezone(zdt::ZonedDateTime, tz::TimeZone) -> ZonedDateTime

Converts a ZonedDateTime from its current TimeZone into the specified TimeZone.

TimeZones.build — Function.

build(version="latest", regions=REGIONS; force=false) -> Void

Builds the TimeZones package with the specified tzdata version and regions. The version is typically a 4-digit year followed by a lowercase ASCII letter (e.g. "2016j"). Users can also specify which regions, or tz source files, should be compiled. Available regions are listed under TimeZones.REGIONS and TimeZones.LEGACY_REGIONS. The force flag is used to re-download tzdata archives.

TimeZones.day

TimeZones.dayofmonth

TimeZones.eval

TimeZones.first_valid — Function.

first_valid(local_dt::DateTime, tz::VariableTimeZone, step::Period)

Construct a valid ZonedDateTime by adjusting the local DateTime. If the local DateTime is non-existent then it will be adjusted using the step to be after the gap. When the local DateTime is ambiguous the first ambiguous DateTime will be returned.

TimeZones.firstdayofmonth

TimeZones.firstdayofquarter

TimeZones.firstdayofweek

TimeZones.firstdayofyear

Dates.guess — Function.

guess(start::ZonedDateTime, finish::ZonedDateTime, step) -> Integer

Given a start and end date, indicates how many steps/periods are between them. Defining this function allows StepRanges to be defined for ZonedDateTimes.

TimeZones.hour

TimeZones.include

TimeZones.interpret — Function.

interpret(dt::DateTime, tz::VariableTimeZone, context::Type{Union{Local,UTC}}) -> Array{ZonedDateTime}

Produces a list of possible ZonedDateTimes given a DateTime and VariableTimeZone. The result will be returned in chronological order. Note that DateTimes in the local context typically return 0-2 results while the UTC context will always return 1 result.

TimeZones.isdst

TimeZones.istimezone — Function.

istimezone(str::AbstractString) -> Bool

Tests whether a string is a valid name for constructing a TimeZone.

TimeZones.julian2zdt

TimeZones.last_valid — Function.

last_valid(local_dt::DateTime, tz::VariableTimeZone, step::Period)

Construct a valid ZonedDateTime by adjusting the local DateTime. If the local DateTime is non-existent then it will be adjusted using the step to be before the gap. When the local DateTime is ambiguous the last ambiguous DateTime will be returned.

TimeZones.lastdayofmonth

TimeZones.lastdayofquarter

TimeZones.lastdayofweek

TimeZones.lastdayofyear

TimeZones.localtime — Function.

localtime(::ZonedDateTime) -> DateTime

Creates a local or civil DateTime from the given ZonedDateTime. For example the 2014-05-30T08:11:24-04:00 would return 2014-05-30T08:11:24.

TimeZones.localzone — Function.

localzone() -> TimeZone

Returns a TimeZone object that is equivalent to the system's current time zone.

TimeZones.millisecond

TimeZones.minute

TimeZones.month

TimeZones.monthday

TimeZones.next_transition_instant — Function.

next_transition_instant(zdt::ZonedDateTime) -> ZonedDateTime
next_transition_instant(tz::TimeZone=localzone()) -> ZonedDateTime

Determine the next instant at which a time zone transition occurs (typically due to daylight-savings time).

Note that the provided ZonedDateTime isn't normally constructable:

julia> instant = next_transition_instant(ZonedDateTime(2018, 3, 1, tz"Europe/London"))
2018-03-25T01:00:00+00:00

julia> instant - Millisecond(1)  # Instant prior to the change
2018-03-25T00:59:59.999+00:00

julia> instant - Millisecond(0)  # Instant after the change
2018-03-25T02:00:00+01:00

julia> ZonedDateTime(2018, 3, 25, 1, tz"Europe/London")  # Cannot normally construct the `instant`
ERROR: NonExistentTimeError: Local DateTime 2018-03-25T01:00:00 does not exist within Europe/London
...

Dates.now — Function.

now(::TimeZone) -> ZonedDateTime

Returns a ZonedDateTime corresponding to the user's system time in the specified TimeZone.

TimeZones.nullable

TimeZones.offset_string

TimeZones.optional

TimeZones.parse_tz_format — Function.

parse_tz_format(str) -> TimeZone

Parse the time zone format typically provided via the "TZ" environment variable. Details on the format can be found under the tzset man page.

Currently this function handles only the first format which is a fixed time zone without daylight saving time.

TimeZones.read_tzfile — Function.

read_tzfile(io::IO, name::AbstractString) -> TimeZone

Read the content of an I/O stream and process it as a POSIX tzfile. The returned TimeZone will be given the suppiled name name unless a FixedTimeZone is returned.

TimeZones.read_tzfile_internal

TimeZones.second

TimeZones.shift_gap — Function.

shift_gap(local_dt::DateTime, tz::VariableTimeZone) -> Array{ZonedDateTime}

Given a non-existent local DateTime in a TimeZone produces two valid ZonedDateTimes that span the gap. Providing a valid local DateTime returns an empty array. Note that this function does not support passing in a UTC DateTime since there are no non-existent UTC DateTimes.

Aside: the function name refers to a period of invalid local time (gap) caused by daylight saving time or offset changes (shift).

TimeZones.show_next_transition — Function.

show_next_transition(io::IO=stdout, zdt::ZonedDateTime)
show_next_transition(io::IO=stdout, tz::TimeZone=localzone())

Display useful information about the next time zone transition (typically due to daylight-savings time). Information displayed includes:

Transition Date: the local date at which the transition occurs (2018-10-28)

Local Time Change: the way the local clock with change (02:00 falls back to 01:00) and the direction of the change ("Forward" or "Backward")

Offset Change: the standard offset and DST offset that occurs before and after the transition

Transition From: the instant before the transition occurs

Transition To: the instant after the transition occurs

julia> show_next_transition(ZonedDateTime(2018, 8, 1, tz"Europe/London"))
Transition Date:   2018-10-28
Local Time Change: 02:00 → 01:00 (Backward)
Offset Change:     UTC+0/+1 → UTC+0/+0
Transition From:   2018-10-28T01:59:59.999+01:00 (BST)
Transition To:     2018-10-28T01:00:00.000+00:00 (GMT)

julia> show_next_transition(ZonedDateTime(2011, 12, 1, tz"Pacific/Apia"))
Transition Date:   2011-12-30
Local Time Change: 00:00 → 00:00 (Forward)
Offset Change:     UTC-11/+1 → UTC+13/+1
Transition From:   2011-12-29T23:59:59.999-10:00
Transition To:     2011-12-31T00:00:00.000+14:00

TimeZones.timezone — Function.

timezone(::ZonedDateTime) -> TimeZone

Returns the TimeZone used by the ZonedDateTime.

TimeZones.timezone_abbrs — Function.

timezone_abbrs -> Array{AbstractString}

Returns a sorted list of all pre-computed time zone abbrevations.

TimeZones.timezone_names — Function.

timezone_names() -> Array{AbstractString}

Returns a sorted list of all of the valid names for constructing a TimeZone.

TimeZones.timezones_from_abbr — Function.

timezones_from_abbr(abbr) -> Array{TimeZone}

Returns all TimeZones that have the specified abbrevation

Dates.today — Function.

today(tz::TimeZone) -> Date

Returns the date portion of now(tz) in local time.

Examples

julia> a, b = now(tz"Pacific/Midway"), now(tz"Pacific/Apia")
(2017-11-09T03:47:04.226-11:00, 2017-11-10T04:47:04.226+14:00)

julia> a - b
0 milliseconds

julia> today(tz"Pacific/Midway"), today(tz"Pacific/Apia")
(2017-11-09, 2017-11-10)

TimeZones.todayat — Function.

todayat(tod::Time, tz::TimeZone, [amb]) -> ZonedDateTime

Creates a ZonedDateTime for today at the specified time of day. If the result is ambiguous in the given TimeZone then amb can be supplied to resolve ambiguity.

Examples

julia> today(tz"Europe/Warsaw")
2017-11-09

julia> todayat(Time(10, 30), tz"Europe/Warsaw")
2017-11-09T10:30:00+01:00

TimeZones.transition_range — Function.

transition_range(dt::DateTime, tz::VariableTimeZone, context::Type{Union{Local,UTC}}) -> UnitRange

Finds the indexes of the tz transitions which may be applicable for the dt. The given DateTime is expected to be local to the time zone or in UTC as specified by context. Note that UTC context will always return a range of length one.

TimeZones.tryparsenext_fixedtz

TimeZones.tryparsenext_tz

TimeZones.unix2zdt

TimeZones.utc — Function.

utc(::ZonedDateTime) -> DateTime

Creates a utc DateTime from the given ZonedDateTime. For example the 2014-05-30T08:11:24-04:00 would return 2014-05-30T12:11:24.

TimeZones.utc_tz

TimeZones.week

TimeZones.year

TimeZones.yearmonth

TimeZones.yearmonthday

TimeZones.zdt2julian

TimeZones.zdt2unix