Docstrings

Reactive.CHANNEL_SIZE

Reactive.MaybeMessage

Reactive.Message

Reactive.NullException

Reactive.Reactive

Reactive.Signal

Reactive.__init__

Reactive._active_binds

Reactive._bindings

Reactive._messages

Reactive.activate!

Reactive.add_action!

Reactive.async_map — Function.

tasks, results = async_map(f, init, input...;typ=typeof(init), onerror=Reactive.print_error)

Spawn a new task to run a function when input signal updates. Returns a signal of tasks and a results signal which updates asynchronously with the results. init will be used as the default value of results. onerror is the callback to be called when an error occurs, by default it is set to a callback which prints the error to stderr. It's the same as the onerror argument to push! but is run in the spawned task.

Reactive.async_mode

Reactive.async_push!

Reactive.auto_name!

Reactive.bind! — Function.

`bind!(dest, src, twoway=true; initial=true)`

for every update to src also update dest with the same value and, if twoway is true, vice-versa. If initial is false, dest will only be updated to src's value when src next updates, otherwise (if initial is true) both dest and src will take src's value immediately.

Reactive.bindmap! — Function.

`bindmap!(dest::Signal, src2dest::Function, src::Signal, dest2src=nothing; initial=true)`

for every update to src also update dest with a modified value (using the function src2dest) and, if dest2src is specified, a two-way update will hold. If initial is false, dest will only be updated to src's modified value when src next updates, otherwise (if initial is true) both dest and src will take their respective modified values immediately.

Reactive.bound_dests — Function.

bound_dests(src::Signal) returns a vector of all signals that will update when src updates, that were bound using bind!(dest, src)

Reactive.bound_srcs — Function.

bound_srcs(dest::Signal) returns a vector of all signals that will cause an update to dest when they update, that were bound using bind!(dest, src)

Reactive.break_loop

Reactive.close

Reactive.connect_delay

Reactive.connect_droprepeats

Reactive.connect_filter

Reactive.connect_filterwhen

Reactive.connect_flatten — Function.

connect_flatten(output, input)

output is the flatten node, input is the Signal{Signal} ("sigsig") node. The flatten needs to update on changes to the input sigsig, or changes to the value of the current sig (current_node). The former is achieved through a foreach wire_flatten attached to the input sigsig. The latter is achieved through binding the flatten to current_node.

Reactive.connect_foldp

Reactive.connect_map

Reactive.connect_merge

Reactive.connect_previous

Reactive.connect_sampleon

Reactive.deactivate!

Reactive.debounce — Function.

debounce(dt, input, f=(acc,x)->x, init=value(input), reinit=x->x;
            typ=typeof(init), name=auto_name!(string("debounce ",dt,"s"), input))

Creates a signal that will delay updating until dt seconds have passed since the last time input has updated. By default, the debounce signal holds the last update of the input signal since the debounce signal last updated.

This behavior can be changed by the f, init and reinit arguments. The init and f functions are similar to init and f in foldp. reinit is called after the debounce sends an update, to reinitialize the initial value for accumulation, it gets one argument, the previous accumulated value.

For example y = debounce(0.2, x, push!, Int[], _->Int[]) will accumulate a vector of updates to the integer signal x and push it after x is inactive (doesn't update) for 0.2 seconds.

Reactive.debug_memory

Reactive.debug_mode

Reactive.delay — Function.

delay(input, default=value(input))

Schedule an update to happen after the current update propagates throughout the signal graph.

Returns the delayed signal.

Reactive.droprepeats — Function.

droprepeats(input)

Drop updates to input whenever the new value is the same as the previous value of the signal.

Reactive.edges

Reactive.eval

Reactive.every — Function.

every(dt)

A signal that updates every dt seconds to the current timestamp. Consider using fpswhen or fps if you want specify the timing signal by frequency, rather than delay.

Reactive.every_connect

Base.filter — Function.

filter(f, default, signal)

remove updates from the signal where f returns false. The filter will hold the value default until f(value(signal)) returns true, when it will be updated to value(signal).

filter(f, signal)

remove updates from the signal where f returns false. The filter will hold the current value of the signal until f(value(signal)) returns true.

Reactive.filterwhen — Function.

filterwhen(switch::Signal{Bool}, default, input)

Keep updates to input only when switch is true.

If switch is false initially, the specified default value is used.

Reactive.flatten — Function.

flatten(input::Signal{Signal}; typ=Any)

Flatten a signal of signals into a signal which holds the value of the current signal. The typ keyword argument specifies the type of the flattened signal. It is Any by default.

Reactive.foldp — Function.

foldp(f, init, inputs...)

Fold over past values.

Accumulate a value as the input signals change. init is the initial value of the accumulator. f should take 1 + length(inputs) arguments: the first is the current accumulated value and the rest are the current input signal values. f will be called when one or more of the inputs updates. It should return the next accumulated value.

Reactive.fps — Function.

fps(rate)

Same as fpswhen(Input(true), rate)

Reactive.fpswhen — Function.

fpswhen(switch, rate)

returns a signal which when switch signal is true, updates rate times every second. If rate is not possible to attain because of slowness in computing dependent signal values, the signal will self adjust to provide the best possible rate.

Reactive.fpswhen_connect

Reactive.getlastactive — Function.

getlastactive(merge_node) Search backwards in nodes, and return the first active node that is one of merge_node's parents

Reactive.include

Reactive.io_lock

Reactive.isactive

Reactive.isnull

Reactive.log_gc

Base.map — Function.

map(f, s::Signal...) -> signal

Transform signal s by applying f to each element. For multiple signal arguments, apply f elementwise.

Reactive.maybe_restart_queue

Base.merge — Function.

merge(inputs...)

Merge many signals into one. Returns a signal which updates when any of the inputs update. If many signals update at the same time, the value of the youngest (most recently created) input signal is taken.

Reactive.node_count

Reactive.nodes

Reactive.nodeset

Reactive.onerror_rethrow

Reactive.pause_push — Function.

Pause a push by recording the active nodes and setting them to inactive. The push can be resumed by reactivating the nodes.

Reactive.preserve — Function.

preserve(signal::Signal)

prevents signal from being garbage collected as long as any of its parents are around. Useful for when you want to do some side effects in a signal. e.g. preserve(map(println, x)) - this will continue to print updates to x, until x goes out of scope. foreach is a shorthand for map with preserve.

Reactive.previous — Function.

previous(input, default=value(input))

Create a signal which holds the previous value of input. You can optionally specify a different initial value.

Reactive.print_error

Reactive.probe

Base.push! — Function.

push!(signal, value, onerror=Reactive.print_error)

Queue an update to a signal. The update will be propagated when all currently queued updates are done processing.

The third (optional) argument, onerror, is a callback triggered when the update ends in an error. The callback receives 4 arguments, onerror(sig, val, node, capex), where sig and val are the Signal and value that push! was called with, respectively, node is the Signal whose action triggered the error, and capex is a CapturedException with the fields ex which is the original exception object, and processed_bt which is the backtrace of the exception.

The default error callback will print the error and backtrace to stderr.

Reactive.reinit_edges!

Reactive.remote_map — Function.

remoterefs, results = remote_map(procid, f, init, input...;typ=typeof(init), onerror=Reactive.print_error)

Spawn a new task on process procid to run a function when input signal updates. Returns a signal of remote refs and a results signal which updates asynchronously with the results. init will be used as the default value of results. onerror is the callback to be called when an error occurs, by default it is set to a callback which prints the error to stderr. It's the same as the onerror argument to push! but is run in the spawned task.

Reactive.remove_action! — Function.

Removes action from node.actions

Reactive.remove_dead_nodes! — Function.

Remove GC'd nodes from nodes, is run before push! when scheduled. Not thread-safe in the sense that if it is run while other code is iterating through nodes, e.g. in run_push, iteration could skip nodes.

Reactive.rename! — Function.

rename!(s::Signal, name::String)

Change a Signal's name

Reactive.restart_queue

Reactive.run — Function.

Processes n messages from the Reactive event queue.

Reactive.run_async

Reactive.run_node — Function.

A node's actions should be run if any of its parents are active, since that generally means one (or more) of the parent nodes' values have changed. If the node doesn't have actions, don't set it to active, since its value won't be updated, meaning the update propagation can stop. N.b. The non-active when node has no actions mechanism is relied on for correct behaviour by fpswhen, and possibly other operators, i.e. it is not just an optimisation.

Reactive.run_push

Reactive.run_remove_dead_nodes

Reactive.run_till_now

Reactive.runaction

Reactive.runner_task

Reactive.sampleon — Function.

sampleon(a, b)

Sample the value of b whenever a updates.

Reactive.schedule_node_cleanup — Function.

Schedule a cleanup of dead nodes - called as a finalizer on each GC'd node

Reactive.set_debug_mode

Reactive.set_value!

Reactive.setup_next_tick

Reactive.stop

Reactive.throttle — Function.

throttle(dt, input, f=(acc,x)->x, init=value(input), reinit=x->x;
            typ=typeof(init), name=auto_name!(string("throttle ",dt,"s"), input), leading=false)

Throttle a signal to update at most once every dt seconds. By default, the throttled signal holds the last update of the input signal during each dt second time window.

This behavior can be changed by the f, init and reinit arguments. The init and f functions are similar to init and f in foldp. reinit is called when a new throttle time window opens to reinitialize the initial value for accumulation, it gets one argument, the previous accumulated value.

For example y = throttle(0.2, x, push!, Int[], _->Int[]) will create vectors of updates to the integer signal x which occur within 0.2 second time windows.

If leading is true, the first update from input will be sent immediately by the throttle signal. If it is false, the first update will happen dt seconds after input's first update

New in v0.4.1: throttle's behaviour from previous versions is now available with the debounce signal type.

Reactive.throttle_connect

Reactive.unbind! — Function.

`unbind!(dest, src, twoway=true)`

remove a link set up using bind!

Reactive.unpreserve — Function.

unpreserve(signal::Signal)

allow signal to be garbage collected. See also preserve.

Reactive.value