Docstrings

CxxWrap.@initcxx — Macro.

@initcxx

Initialize the C++ pointer tables in a precompiled module using CxxWrap. Must be called from within __init__ in the wrapped module

CxxWrap.@readmodule — Macro.

@readmodule libraryfile [functionname]

Read a C++ module and associate it with the Julia module enclosing the macro call.

CxxWrap.@safe_cfunction

CxxWrap.@wrapfunctions — Macro.

@wrapfunctions

Wrap the functions defined in the C++ side of the enclosing module. Requires that @readmodule and @wraptypes was called first.

CxxWrap.@wrapmodule — Macro.

@wrapmodule libraryfile [functionname]

Place the functions and types from the C++ lib into the module enclosing this macro call Calls an entry point named define_julia_module, unless another name is specified as the second argument.

CxxWrap.@wraptypes — Macro.

@wraptypes

Wrap the types defined in the C++ side of the enclosing module. Requires that @readmodule was called first.

CxxWrap.CallOpOverload

CxxWrap.ConstArray

CxxWrap.ConstPtr

CxxWrap.ConstructorFname

CxxWrap.CppEnum

CxxWrap.CppFunctionInfo

CxxWrap.CxxWrap

CxxWrap.IsCxxType

CxxWrap.IsNormalType

CxxWrap.SafeCFunction

CxxWrap.SmartPointer — Type.

Base class for smart pointers

CxxWrap.SmartPointerWithDeref — Type.

Concrete smart pointer implementation. PT is a hash for the pointer type.

CxxWrap.StrictlyTypedNumber

CxxWrap.__cxxwrap_smartptr_cast_to_base

CxxWrap.__cxxwrap_smartptr_construct_from_other

CxxWrap.__cxxwrap_smartptr_dereference

CxxWrap.__excluded_names

CxxWrap.__init__

CxxWrap.allocated_types

CxxWrap.argument_overloads

CxxWrap.bind_constants

CxxWrap.build_function_expression

CxxWrap.check_deps

CxxWrap.cpp_trait_type

CxxWrap.cxxdowncast

CxxWrap.depsfile

CxxWrap.eval

CxxWrap.gcprotect — Function.

Protect a variable from garbage collection by adding it to the global array kept by CxxWrap

CxxWrap.gcunprotect — Function.

Unprotect a variable from garbage collection by removing it from the global array kept by CxxWrap

CxxWrap.get_module_functions

CxxWrap.has_cxx_module

CxxWrap.include

CxxWrap.isnull

CxxWrap.jlcxx_path

CxxWrap.lib_path

CxxWrap.libcxxwrap_julia

CxxWrap.load_module

CxxWrap.make_func_declaration

CxxWrap.map_julia_arg_type

CxxWrap.prefix

CxxWrap.process_fname

CxxWrap.ptrunion — Function.

Create a Union containing the type and a smart pointer to any type derived from it

CxxWrap.readmodule

CxxWrap.reference_type

CxxWrap.reference_types

CxxWrap.register_julia_module

CxxWrap.smart_pointer_type

CxxWrap.wrap_functions

CxxWrap.wrap_reference_converters

CxxWrap.wrapfunctions

CxxWrap.wrapmodule

CxxWrap.wraptypes

CxxWrap.wstring_to_cpp

CxxWrap.wstring_to_julia