Docstrings

WriteVTK.CollectionFile

WriteVTK.DEFAULT_COMPRESSION_LEVEL

WriteVTK.DatasetFile

WriteVTK.IS_LITTLE_ENDIAN

WriteVTK.InputDataType — Constant.

Types allowed as input to vtk_point_data() and vtk_cell_data().

Either (abstract) arrays or tuples of arrays are allowed. In the second case, the length of the tuple determines the number of components of the input data (e.g. if N = 3, it corresponds to a 3D vector field).

WriteVTK.MeshCell

WriteVTK.MultiblockFile

WriteVTK.VTKCellType

WriteVTK.VTKCellTypes

WriteVTK.VTKDataType — Constant.

Union of data types allowed by VTK (see file-formats.pdf, page 15).

WriteVTK.VTKFile

WriteVTK.WriteVTK

WriteVTK._compression_level

WriteVTK._dep_message_VTKCellType

WriteVTK.add_extension

WriteVTK.collection_add_timestep

WriteVTK.data_to_xml — Function.

Add numerical data to VTK XML file.

Data is written under the xParent XML node.

Nc corresponds to the number of components of the data.

WriteVTK.data_to_xml_appended — Function.

Add appended raw binary data to VTK XML file.

Data is written to the vtk.buf buffer.

When compression is enabled:

the data array is written in compressed form (obviously);

the header, written before the actual numerical data, is an array of UInt32 values: [num_blocks, blocksize, last_blocksize, compressed_blocksizes] All the sizes are in bytes. The header itself is not compressed, only the data is. For more details, see: http://public.kitware.com/pipermail/paraview/2005-April/001391.html http://mathema.tician.de/what-they-dont-tell-you-about-vtk-xml-binary-formats (This is not really documented in the VTK specification...)

Otherwise, if compression is disabled, the header is just a single UInt32 value containing the size of the data array in bytes.

WriteVTK.data_to_xml_inline — Function.

Add inline, base64-encoded data to VTK XML file.

WriteVTK.datatype_str — Function.

Return the VTK string representation of a numerical data type.

WriteVTK.eval

WriteVTK.extent_attribute — Function.

Return the "extent" attribute required for structured (including rectilinear) grids.

WriteVTK.include

WriteVTK.multiblock_add_block

WriteVTK.num_cells_structured — Function.

Number of cells in structured grids.

In 3D, all cells are hexahedrons (i.e. VTKHEXAHEDRON), and the number of cells is (Ni-1)(Nj-1)(Nk-1). In 2D, they are quadrilaterals (VTKQUAD), and in 1D they are line segments (VTK_LINE).

WriteVTK.num_components — Function.

Determine number of components of input data.

WriteVTK.paraview_collection

WriteVTK.rectilinear_grid

WriteVTK.save_with_appended_data — Function.

Write VTK XML file containing appended binary data to disk.

In this case, the XML file is written manually instead of using the save_file function of LightXML, which doesn't allow to write raw binary data.

WriteVTK.sizeof_data — Function.

Total size of data in bytes.

WriteVTK.structured_grid

WriteVTK.unstructured_grid

WriteVTK.vtk_cell_data

WriteVTK.vtk_grid

WriteVTK.vtk_multiblock

WriteVTK.vtk_point_data

WriteVTK.vtk_point_or_cell_data — Function.

Add either point or cell data to VTK file.

Here Nc is the number of components of the data (Nc >= 1).

WriteVTK.vtk_save

WriteVTK.vtk_write_array — Function.

Write a Julia array to a VTK image data file (.vti).

Useful for general visualisation of arrays. The input can be a 2D or 3D array.

WriteVTK.vtk_xml_write_header

WriteVTK.write_array — Function.

Write array of numerical data to stream.