Docstrings

MatrixDepot.@addgroup — Macro.

add a group to Matrix Depot

MatrixDepot.@modifygroup — Macro.

add or replace group in Matrix Depot

MatrixDepot.@pred — Macro.

@pred(expression)

Generate a predicate function using the expression as function body. Variable names within the expression, which are properties of RemoteMatrixData (e.g. title, m, nnz) are used to access data.title etc. Other variable names, are used from the outer scope.

example: maxnnz = 1_000; listnames(@pred(n <= maxnnz)) would produce a list of all data with less than maxnnz structural non-zeros.

MatrixDepot.@rmgroup — Macro.

remove an added group from Matrix Depot

MatrixDepot.ALL_PATTERN

MatrixDepot.ARRAY

MatrixDepot.AbstractNot

MatrixDepot.Alias

MatrixDepot.Auxiliar

MatrixDepot.COMPLEX

MatrixDepot.COORD

MatrixDepot.DATA_DIR

MatrixDepot.DataError

MatrixDepot.EMPTY_PATTERN

MatrixDepot.FunctionName

MatrixDepot.GENERAL

MatrixDepot.GeneratedMatrixData

MatrixDepot.Group

MatrixDepot.HERMITIAN

MatrixDepot.INTEGER

MatrixDepot.IntOrVec

MatrixDepot.MATRIX

MatrixDepot.MATRIXCLASS — Constant.

predefined matrix classes (for the generated functions)

MatrixDepot.MATRIXDICT — Constant.

Associate names with matrix-generating functions

MatrixDepot.MATRIXM

MatrixDepot.MATRIX_DB — Constant.

The place to store all matrix data in process

MatrixDepot.MD

MatrixDepot.MMField

MatrixDepot.MMFieldComplex

MatrixDepot.MMFieldInteger

MatrixDepot.MMFieldPattern

MatrixDepot.MMFieldReal

MatrixDepot.MMFormat

MatrixDepot.MMFormatArray

MatrixDepot.MMFormatCoordinate

MatrixDepot.MMObject

MatrixDepot.MMObjectMatrix

MatrixDepot.MMProperties

MatrixDepot.MMProperty

MatrixDepot.MMRemoteType

MatrixDepot.MMSymmetry

MatrixDepot.MMSymmetryGeneral

MatrixDepot.MMSymmetryHermitian

MatrixDepot.MMSymmetrySkewSymmetric

MatrixDepot.MMSymmetrySymmetric

MatrixDepot.MM_NAME_TO_PROP

MatrixDepot.MM_REMOTE

MatrixDepot.MY_DEPOT_DIR

MatrixDepot.MatrixData

MatrixDepot.MatrixDatabase

MatrixDepot.MatrixDepot — Module.

julia MatrixDepot

Give access to a wealth of sample and test matrices and accompanying data. A set of matrices is generated locally (with arguments controlling the special case). Another set is loaded from one of the publicly accessible matrix collections SuiteSparse Matrix Collection (formerly University of Florida Matrix Collection) and the Matrix Market Collection.

Access is like

using MatrixDepot

A = matrixdepot("hilb", 10) # locally generated hilbert matrix dimensions (10,10)

A = matrixdepot("HB/1138_bus")     # named matrix of the SuiteSparse Collection
A = matrixdepot(sp(1))             # same matrix using numerical id
A = matrixdepot("Harwell*/*/1138_bus") # matrix from the Matrix Market Collection 

md = mdopen("*/bfly")   # named matrix with some extra data
A = md.A
co = md.coord
tx = md("Gname_10.txt")

md = mdopen("gravity", 10, false) # localy generated example with rhs and solution
A = md.A
b = md.b
x = md.x

commands:

mdinfo, listdir, listgroups, matrixdepot, mdopen, listdata, mdlist,
metasymbols, @addgroup, @modifygroup, @rmgroup.

selector patterns:

strings, string-patterns (using "*", "?", "/", "**"), regular expressions: for names
builtin(42), user(3,5), sp(10:11,6,2833), mm(1): to access by integer id

predicate patterns:

isboolean, isinteger, isreal, iscomplex
isgeneral, issymmetric, ishermitian, isskew
isbuiltin, isuser, islocal, isremote, isloaded, isunloaded
keyword(string expression), logical, hasdata(symbol), @pred(expression)

see also: "logical" for logical combinations of all kinds of patterns.

MatrixDepot.MatrixDescriptor — Type.

MatrixDescriptor{T<:MatrixData}

Access object which is created by a call to mdopen(::Pattern). Several user functions allow to access data according to the unique pattern. Keeps data cache for exactly the same calling arguments (in the case of generated objects). For remote objects the data cache keeps all available metadata.

MatrixDepot.MatrixGenerator

MatrixDepot.MetaInfo

MatrixDepot.Not

MatrixDepot.PATTERN

MatrixDepot.P_block

MatrixDepot.Pattern

MatrixDepot.REAL

MatrixDepot.RegProb

MatrixDepot.RegProbNoSolution

MatrixDepot.RemoteMatrixData

MatrixDepot.RemoteParameters

MatrixDepot.RemoteType

MatrixDepot.SKEW_SYMMETRIC

MatrixDepot.SUBSETS

MatrixDepot.SYMMETRIC

MatrixDepot.TA_REMOTE

MatrixDepot.TURemoteType

MatrixDepot.UF_REMOTE

MatrixDepot.USERMATRIXDICT — Constant.

user-defined matrix generators
to be populated with `include_generator` in user source code

MatrixDepot.__init__

MatrixDepot._downloadindices

MatrixDepot._flatten_pattern

MatrixDepot._isfield

MatrixDepot._issymmetry

MatrixDepot._load

MatrixDepot._mdheader

MatrixDepot._mdopen

MatrixDepot._metareader

MatrixDepot._parsenext

MatrixDepot._repl

MatrixDepot.addmetadata!

MatrixDepot.addtogroup

MatrixDepot.aliasname — Function.

aliasname(data::MatrixData)
aliasname(Type{<:MatrixData, id::Integer)

return alias name derived from integer id

MatrixDepot.aliasresolve

MatrixDepot.alternate

MatrixDepot.argerr

MatrixDepot.baart — Function.

Fredholm Integral Equation of the First Kind

Input options:

[type,] dim, [matrixonly]: the dimenstion of the matrix is dim. If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated (matrixonly = true by default).

Groups: ["regprob"]

References:

M. L. Baart, The use of auto-correlation for pesudo-rank determination in noisy ill-conditioned linear-squares problems, IMA, J. Numer. Anal. 2 (1982), 241-247.

MatrixDepot.binomialm — Function.

Binomial Matrix

The matrix is a multiple of an involutory matrix.

Input options:

[type,] dim: the dimension of the matrix.

MatrixDepot.blur — Function.

Image Deblurring Test Problem

The generated matrix A is an n*n-by-n*n sparse, symmetric, doubly block Toeplitz matrix that models blurring of an n-by-n image by a Gaussian point spread function.

Input options:

[type,] dim, band, σ, [matrixonly]: the dimension of the matrix is dim^2. band is the half-bandwidth, only matrix elements within a distance band-1 from the diagonal are nonzero. σ controls the width of the Gaussin point spread function. The larger the σ, the wider the function and the more ill posed the problem. If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated (matrixonly = true by default).

[type,] dim, [matrixonly]: band = 3, σ = 0.7.

Groups: ["regprob", "sparse"]

MatrixDepot.buildnametable — Function.

format output list to multi- column table form adapted to display width.

MatrixDepot.buildnametable1

MatrixDepot.builtin

MatrixDepot.cauchy — Function.

Cauchy Matrix

Given two vectors x and y, the (i,j) entry of the Cauchy matrix is 1/(x[i]+y[j]).

Input options:

[type,] x, y: two vectors.

[type,] x: a vector. y defaults to x.

[type,] dim: the dimension of the matrix. x and y default to [1:dim;].

Groups: ["inverse", "illcond", "symmetric", "posdef"]

References:

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.1

MatrixDepot.chebspec — Function.

Chebyshev Spectral Differentiation Matrix

If k = 0,the generated matrix is nilpotent and a vector with all one entries is a null vector. If k = 1, the generated matrix is nonsingular and well-conditioned. Its eigenvalues have negative real parts.

Input options:

[type,] dim, k: dim is the dimension of the matrix and k = 0 or 1.

[type,] dim: k=0.

Groups: ["eigen"]

References:

L. N. Trefethen and M. R. Trummer, An instability phenomenon in spectral methods, SIAM J. Numer. Anal., 24 (1987), pp. 1008-1023.

MatrixDepot.check_symbols — Function.

check_symbols(p::Pattern)

throw ArgumentError if pattern uses unknown symbol as a group name.

MatrixDepot.chow — Function.

Chow Matrix

The Chow matrix is a singular Toeplitz lower Hessenberg matrix.

Input options:

[type,] dim, alpha, delta: dim is dimension of the matrix. alpha, delta are scalars such that A[i,i] = alpha + delta and A[i,j] = alpha^(i + 1 -j) for j + 1 <= i.

[type,] dim: alpha = 1, delta = 0.

Groups: ["eigen"]

References:

T. S. Chow, A class of Hessenberg matrices with known eigenvalues and inverses, SIAM Review, 11 (1969), pp. 391-395.

MatrixDepot.circul — Function.

Circulant Matrix

A circulant matrix has the property that each row is obtained by cyclically permuting the entries of the previous row one step forward.

Input options:

[type,] vec, col_dim: a vector and the column dimension.

[type,] vec: a vector.

[type,] dim: the dimension of the matrix.

Groups: ["symmetric", "posdef", "eigen"]

References:

P. J. Davis, Circulant Matrices, John Wiley, 1977.

MatrixDepot.clement — Function.

Clement Matrix

The Clement matrix is a tridiagonal matrix with zero diagonal entries. If k = 1, the matrix is symmetric.

Input options:

[type,] dim, k: dim is the dimension of the matrix. If k = 0, the matrix is of type Tridiagonal. If k = 1, the matrix is of type SymTridiagonal.

[type,] dim: k = 0.

Groups: ["inverse", "symmetric", "eigen"]

References:

P. A. Clement, A class of triple-diagonal matrices for test purposes, SIAM Review, 1 (1959), pp. 50-52.

MatrixDepot.colval — Function.

colval(A)

reconstruct colum-indices from colptr of SparseMatrixCSC.

MatrixDepot.companion — Function.

Companion Matrix

The companion matrix to a monic polynomial a(x) = a_0 + a_1x + ... + a_{n-1}x^{n-1} + x^n is the n-by-n matrix with ones on the subdiagonal and the last column given by the coefficients of a(x).

Input options:

[type,] vec: vec is a vector of coefficients.

[type,] dim: vec = [1:dim;]. dim is the dimension of the matrix.

MatrixDepot.data_warn

MatrixDepot.dataurl

MatrixDepot.daterr

MatrixDepot.deriv2 — Function.

Computation of the Second Derivative

A classical test problem for regularization algorithms: This is a mildly ill-posed problem. It is a discretization of a first kind Fredholm integral equation whose kernel K is the Green's function for the second derivative.

Input options:

[type,] dim, example, [matrixonly]: the dimension of the matrix is dim. One choose between between the following right-hand g and solution f:

    example = 1 gives g(s) = (s^3 - s)/6, f(t) = t.

    example = 2 gives g(s) = exp(s) + (1 -e)s - 1, f(t) = exp(t)

    example = 3 gives  g(s) = | (4s^3 - 3s)/24,  s < 0.5
                              | (-4s^3 + 12s^2 - 9s + 1)/24, s>= 0.5
                       f(t) = | t, t < 0.5
                       f(t) = | 1- t, t >= 0.5.

    If `matrixonly = false`, the matrix A and vectors b and x in the linear system Ax = b will be
    generated (`matrixonly = true` by default).

[type,] dim, [matrixonly]: example = 1.

Groups: ["regprob"]

References:

P. C. Hansen, Regularization tools: A MATLAB pacakge for analysis and solution of discrete ill-posed problems. Numerical Algorithms, 6(1994), pp.1-35

MatrixDepot.dingdong — Function.

Dingdong Matrix

The Dingdong matrix is a symmetric Hankel matrix invented by DR. F. N. Ris of IBM, Thomas J Watson Research Centre. The eigenvalues cluster around π/2 and -π/2.

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["symmetric", "eigen"]

References:

J. C. Nash, Compact Numerical Methods for Computers: Linear Algebra and Function Minimisation, second edition, Adam Hilger, Bristol, 1990 (Appendix 1).

MatrixDepot.downloadindex

MatrixDepot.downloadindices — Function.

MatrixDepot.downloadindices(db)

download html files and store matrix data in db. additionally generate aliases for local and loaded matrices. Serialize db object in file db.data. If a file db.data is present in the data directory, deserialize the data instead of downloading and collection data.

MatrixDepot.erdrey — Function.

Erdos-Renyi Random Graph

Generate an adjacency matrix of an Edros-Renyi random graph: an undirected graph is chosen uniformly at random from the set of all symmetric graphs with n nodes and m edges.

Input options:

[type,] n, m: the dimension of the matrix is n. The number of 1's in the matrix is 2*m.

[type,] n: m = ceil(Int, n*log(n)/2).

Groups: ["sparse", "graph"]

References:

P. Erdos and A. Renyi, On Random Graphs, Publ. Math. Debrecen, 6, 1959, pp. 290-297

MatrixDepot.eval

MatrixDepot.extension

MatrixDepot.extract_names

MatrixDepot.extract_symbols

MatrixDepot.extremnnz

MatrixDepot.fiedler — Function.

Fiedler Matrix

The Fiedler matrix is symmetric matrix with a dominant positive eigenvalue and all the other eigenvalues are negative.

Input options:

[type,] vec: a vector.

[type,] dim: dim is the dimension of the matrix. vec=[1:dim;].

*Groups: *["inverse", "symmetric", "eigen"]

References:

G. Szego, Solution to problem 3705, Amer. Math. Monthly, 43 (1936), pp. 246-259.

J. Todd, Basic Numerical Mathematics, Vol. 2: Numerical Algebra, Birkhauser, Basel, and Academic Press, New York, 1977, p. 159.

MatrixDepot.filename

MatrixDepot.fillcache!

MatrixDepot.flatten_pattern — Function.

flattenPattern(p::Pattern)

return the vector of all elementary patterns, contained in the pattern.

MatrixDepot.fname — Function.

return the name of the function f as a string.

MatrixDepot.forsythe — Function.

Forsythe Matrix

The Forsythe matrix is a n-by-n perturbed Jordan block. This generator is adapted from N. J. Higham's Test Matrix Toolbox.

Input options:

[type,] dim, alpha, lambda: dim is the dimension of the matrix. alpha and lambda are scalars.

[type,] dim: alpha = sqrt(eps(type)) and lambda = 0.

Groups: ["inverse", "illcond", "eigen"]

MatrixDepot.foxgood — Function.

Severely Ill-posed Problem Suggested by Fox & Goodwin

This is a model problem discretized by simple quadrature, which does not satifiy the discrete Picard condition for the small singular values.

Input options:

[type,] dim, [matrixonly]: dim is the dimension of the matrix. If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated (matrixonly = true by default).

Groups: ["regprob"]

References:

C. T. H. Baker, The Numerical Treatment of Integral Equations, Clarendon Press, Oxford, 1977, p. 665.

MatrixDepot.frank — Function.

Frank Matrix

The Frank matrix is an upper Hessenberg matrix with determinant 1. The eigenvalues are real, positive and very ill conditioned.

Input options:

[type,] dim, k: dim is the dimension of the matrix, k = 0 or 1. If k = 1 the matrix reflect about the anti-diagonal.

[type,] dim: the dimension of the matrix.

Groups: ["illcond", "eigen"]

References:

W. L. Frank, Computing eigenvalues of complex matrices by determinant evaluation and by methods of Danilewski and Wielandt, J. Soc. Indust. Appl. Math., 6 (1958), pp. 378-392 (see pp. 385, 388).

MatrixDepot.geo

MatrixDepot.getbytes

MatrixDepot.gilbert — Function.

Gilbert Random Graph

Generate an adjecency matrix of a Gilbert random graph: an undirected graph with pairs of nodes are connected with indepdent probability p.

Input options:

[type,] n, p: the dimension of the matrix is n and the probability that any two nodes are connected is p.

[type,] n: p = log(n)/n.

Groups: ["sparse", "graph"]

References:

E.N. Gilbert, Random Graphs, Ann. Math. Statist., 30, (1959) pp. 1141-1144.

MatrixDepot.golub — Function.

Golub Matrix

Golub matrix is the product of two random unit lower and upper triangular matrices respectively. LU factorization without pivoting fails to reveal that such matrices are badly conditioned.

Input options:

[type,] dim: the dimension of the matrix.

References:

D. Viswanath and N. Trefethen. Condition Numbers of Random Triangular Matrices, SIAM J. Matrix Anal. Appl. 19, 564-581, 1998.

MatrixDepot.gravity — Function.

One-dimensional Gravity Surverying Problem

Discretization of a 1-D model problem in gravity surveying, in which a mass distribution f(t) is located at depth d, while the vertical component of the gravity field g(s) is measured at the surface.

Input options:

[type,] dim, example, a, b, d, [matrixonly]: dim is the dimension of the matrix. Three examples are implemented.

   (a) example = 1 gives f(t) = sin(π*t) + 0.5*sin(2*π*t).

   (b) example = 2 gives f(t) = piecewise linear function.

   (c) example = 3 gives f(t) = piecewise constant function.

   The t integration interval is fixed to [0, 1], while the s
   integration interval [a, b] can be specified by the user.
   The parameter d is the depth at which the magnetic deposit is
   located. The larger the d, the faster the decay of the singular values.

   If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated
   (matrixonly = true by default).

[type,] dim, example, [matrixonly]: a = 0, b = 1, d = 0.25;

[type,] dim, [matrixonly]: example = 1, a = 0, b = 1, d = 0.25.

Groups: ["regprob"]

References:

G. M. Wing and J. D. Zahrt, A Primer on Integral Equations of the First Kind, Society for Industrial and Applied Mathematics, Philadelphia, 1991, p. 17.

MatrixDepot.grcar — Function.

Grcar Matrix

The Grcar matrix is a Toeplitz matrix with sensitive eigenvalues.

Input options:

[type,] dim, k: dim is the dimension of the matrix and k is the number of superdiagonals.

[type,] dim: the dimension of the matrix.

Groups: ["eigen"]

References:

J. F. Grcar, Operator coefficient methods for linear equations, Report SAND89-8691, Sandia National Laboratories, Albuquerque, New Mexico, 1989 (Appendix 2).

MatrixDepot.gunzip

MatrixDepot.hadamard — Function.

Hadamard Matrix

The Hadamard matrix is a square matrix whose entries are 1 or -1. It was named after Jacques Hadamard. The rows of a Hadamard matrix are orthogonal.

Input options:

[type,] dim: the dimension of the matrix, dim is a power of 2.

Groups: ["inverse", "orthogonal", "eigen"]

References:

S. W. Golomb and L. D. Baumert, The search for Hadamard matrices, Amer. Math. Monthly, 70 (1963) pp. 12-17

MatrixDepot.hankel — Function.

Hankel Matrix

A Hankel matrix is a matrix that is symmetric and constant across the anti-diagonals.

Input options:

[type,] vc, vr: vc and vc are the first column and last row of the matrix. If the last element of vc differs from the first element of vr, the last element of rc prevails.

[type,] v: vc = vr = v.

[type,] dim: dim is the dimension of the matrix. v = [1:dim;].

MatrixDepot.hasdata — Function.

hasdata(meta::Union{Symbol,Tuple,Vector})

Predicate function checks, if matrix data have metadata symbol meta. Tuples and Vectors are interpreted as AND resp. OR.

MatrixDepot.hasinfo

MatrixDepot.heat — Function.

Inverse Heat Equation

Input options:

[type,] dim, κ, [matrixonly]: dim is the dimension of the matrix and dim must be even. κ controls the ill-conditioning of the matrix. (κ = 5 gives a well-conditioned problem and κ = 1 gives an ill conditoned problem). If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated (matrixonly = true by default).

[type,] n, [matrixonly]: κ = 1.

Groups: ["regprob"]

References:

A. S. Carasso, Determining surface temperatures from interior observations, SIAM J. Appl. Math. 42 (1982), 558-574.

MatrixDepot.hilb — Function.

Hilbert Matrix

The Hilbert matrix has (i,j) element 1/(i+j-1). It is notorious for being ill conditioned. It is symmetric positive definite and totally positive.

Input options:

[type,] dim: the dimension of the matrix.

[type,] rowdim, coldim: the row and column dimensions.

Groups: ["inverse", "illcond", "symmetric", "posdef"]

References:

M. D. Choi, Tricks or treats with the Hilbert matrix, Amer. Math. Monthly, 90 (1983), pp. 301-312.

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.1.

MatrixDepot.ilaplace — Function.

Inverse Laplace Transformation

MatrixDepot.include

MatrixDepot.include_generator

MatrixDepot.indexurl

MatrixDepot.init

MatrixDepot.insertlocal

MatrixDepot.invhilb — Function.

Inverse of the Hilbert Matrix

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["inverse", "illcond", "symmetric","posdef"]

References:

M. D. Choi, Tricks or treats with the Hilbert matrix, Amer. Math. Monthly, 90 (1983), pp. 301-312.

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.1.

MatrixDepot.invol — Function.

Involutory Matrix

An involutory matrix is a matrix that is its own inverse.

Input options:

[type,] dim: dim is the dimension of the matrix.

Groups: ["inverse", "illcond", "eigen"]

References:

A. S. Householder and J. A. Carpenter, The singular values of involutory and of idempotent matrices, Numer. Math. 5 (1963), pp. 234-237.

MatrixDepot.is_all

MatrixDepot.is_pure_string

MatrixDepot.is_pure_vector

MatrixDepot.isboolean

MatrixDepot.isbuiltin

MatrixDepot.iscomplex

MatrixDepot.isgeneral

MatrixDepot.ishermitian

MatrixDepot.isinteger

MatrixDepot.isloaded

MatrixDepot.islocal

MatrixDepot.isreal

MatrixDepot.isremote

MatrixDepot.isskew

MatrixDepot.issymmetric

MatrixDepot.isunloaded

MatrixDepot.isuser

MatrixDepot.jordbloc

MatrixDepot.kahan — Function.

Kahan Matrix

The Kahan matrix is an upper trapezoidal matrix, i.e., the (i,j) element is equal to 0 if i > j. The useful range of θ is 0 < θ < π. The diagonal is perturbed by pert*eps()*diagm([n:-1:1;]).

Input options:

[type,] rowdim, coldim, θ, pert: rowdim and coldim are the row and column dimensions of the matrix. θ and pert are scalars.

[type,] dim, θ, pert: dim is the dimension of the matrix.

[type,] dim: θ = 1.2, pert = 25.

Groups: ["inverse", "illcond"]

References:

W. Kahan, Numerical linear algebra, Canadian Math. Bulletin, 9 (1966), pp. 757-801.

MatrixDepot.keyfor

MatrixDepot.keyword — Function.

keyword(word::Union{AbstractString,Tuple,Vector})

Predicate function checks, if word is contained in on to the textual metadata fields [:notes, :title, :kind, :author]. Tuples and Vectors are interpreted as AND resp. OR.

MatrixDepot.kms — Function.

Kac-Murdock-Szego Toeplitz matrix

Input options:

[type,] dim, rho: dim is the dimension of the matrix, rho is a scalar such that A[i,j] = rho^(abs(i-j)).

[type,] dim: rho = 0.5.

Groups: ["inverse", "illcond", "symmetric", "posdef"]

References:

W. F. Trench, Numerical solution of the eigenvalue problem for Hermitian Toeplitz matrices, SIAM J. Matrix Analysis and Appl., 10 (1989), pp. 135-146 (and see the references therein).

MatrixDepot.lehmer — Function.

Lehmer Matrix

The Lehmer matrix is a symmetric positive definite matrix. It is totally nonnegative. The inverse is tridiagonal and explicitly known

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["inverse", "symmetric", "posdef"]

References:

M. Newman and J. Todd, The evaluation of matrix inversion programs, J. Soc. Indust. Appl. Math., 6 (1958), pp. 466-476. Solutions to problem E710 (proposed by D.H. Lehmer): The inverse of a matrix, Amer. Math. Monthly, 53 (1946), pp. 534-535.

MatrixDepot.list!

MatrixDepot.list_all

MatrixDepot.list_builtin

MatrixDepot.list_local

MatrixDepot.list_user

MatrixDepot.listdata — Function.

listdata([db,] p::Pattern)

Return an array of MatrixData objects according to matched patterns.

MatrixDepot.listdir — Function.

listdir([db,] p::AbstractString)

list directories and the number of matrices contained in them. get an overview of the count of names down directories. return a list with summary information for directories in matrix name space. The input argument is split into 2 patterns by the first double slash "//". The whole string (with multiple slashes reduced to single slashes) determines a subset of all matrix names. They are then grouped by the first pattern and for each different group value the number of names in the subset is counted. A final / is replaced by "//**".

E.g.

listdir("/*") - count names without a /.

listdir("/k*") - count names without / starting with k*.

listdir("*//*") - count names with one directory part (sp-collection)

listdir("*/*//*") - count names with two directory parts (mm-collection)

listdir("*//*/*") - count names with two directory parts (mm-collection)

listdir("Har*//*/*") - restrict to directories starting with "Har"

listdir("Har*/*//*") - all subdirectoreis of the previous

MatrixDepot.listgroups — Function.

listgroups()

Return a list of available groups (array of Symbol).

MatrixDepot.listnames

MatrixDepot.load — Function.

load([db,] pattern)

Load data from remote repository for all problems matching pattern.

Return the number of successfully loaded matrices.

MatrixDepot.loadinfo — Function.

loadinfo([db,] pattern)

loadinfo(data::RemoteDate)

Download the first part of the data file. Stop reading, as soon as the initial comment an the size values of the main matrix have been finished. Stor this in a file with extension .info in the same directory, where the .mtx file is. If the complete file is already availble, the download is not performed, because the head of the .mtx file contains the same lines.

MatrixDepot.loadmatrix — Function.

loadmatrix(data::RemoteMatrixData)

Download the files backing the data from a remote repository. That is currently the TAMU site for the UFl collection and the NIST site for the MatrixMarket caoolection. The files are uncompressed and un-tar-ed if necessary. The data files containing the matrix data have to be in MatrixMarket format in both cases. Note, that some of the files of the MM collection are not available in MatrixMarket format. An error message results, if trie to load them.

MatrixDepot.localbase

MatrixDepot.localdir

MatrixDepot.localfile

MatrixDepot.localindex

MatrixDepot.logical — Function.

Pattern syntactic sugar

the logical operators |, &, and ~ may be applied to all kind of Patterns with the usual meaning.

~ : unary negation operator - pattern does not match (highest priority)

& : binary logical and - both patterns match

| : binary logical or - any of the pattern match (lowest priority)

parentheses can be used to overrule operator precedence.

[p...] is the same as p[1] | p[2] ...

(p...) is the same as p[1] & p[2] ...

~(p...) === ~((p...)) - that is ~(p[1] & p[2] ...)

Precedence of '*' is higher that ~ for character and string objects:

so ~ "a" * "b" === ~("a" * "b") === ~"ab" also ~'a'^2*"b" === ~"aab"

MatrixDepot.lotkin — Function.

Lotkin Matrix

The Lotkin matrix is the Hilbert matrix with its first row altered to all ones. It is unsymmetric, illcond and has many negative eigenvalues of small magnitude.

Input options:

[type,] dim: dim is the dimension of the matrix.

Groups: ["inverse", "illcond", "eigen"]

References:

M. Lotkin, A set of test matrices, MTAC, 9 (1955), pp. 153-161.

MatrixDepot.madjoint — Function.

madjoint(A)

Materialized adjoint of sparse Matrix

MatrixDepot.magic — Function.

Magic Square Matrix

The magic matrix is a matrix with integer entries such that the row elements, column elements, diagonal elements and anti-diagonal elements all add up to the same number.

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["inverse"]

MatrixDepot.make_func

MatrixDepot.make_pred

MatrixDepot.matrixdepot — Function.

matrixdepot(p::Pattern, args...)

Return matrix according to pattern or local matrix according to name and arguments.

If not loaded, load remote matrix first. p must be a unique pattern (match only one name). The presence of arguments makes sense only if the pattern matches the name of a generated (=local) matrix.

Only the matrix part is delivered, also in the local cases, where the underlying function returns a structure containing matrix and vectors. Use md = mdopen; md.A, md.b ... to access those objects.

MatrixDepot.matrixfile

MatrixDepot.matrixinfofile

MatrixDepot.mdata — Function.

mdata(db, pattern)

Return unique MatrixData object according to pattern.

MatrixDepot.mdatav

MatrixDepot.mdinfo — Function.

return information about all matrices selected by pattern

mdinfo()

Overview about matrices.

MatrixDepot.mdlist — Function.

mdlist([db,] p::Pattern)

return a vector of full matrix names where name or alias match given pattern. p can be one of the following:

a plain string (without characters * and ?) which must match exactly

a string containing * and ? acting like a shell path pattern

a regular expression

an integer matching equivalent to the alias string "#$p"

a range of integers

a group name expressed as a symbol e.g. :local, :all, :illcond, posdef

the name of a predicate function f(::MatrixData)::Bool, e.g. symmetric, ...

a vector of patterns meaning the union

a tuple of patterns meaning the intersection

MatrixDepot.mdopen — Function.

mdopen([db,] pattern)
mdopen(f, [db,] pattern)

Return MatrixDescriptor object, which can be used with data access functions.

Make sure that data files are loaded. Keeps a cache of already delivered matrices and metadata. If the pattern has not a unique resolution, an error is thrown.

MatrixDepot.meshgrid

MatrixDepot.metadata — Function.

metadata([db, ], Union{MatrixDescriptor,MatrixData})

Return copy of list of metadata names.

MatrixDepot.metaname

MatrixDepot.metareader — Function.

metareader(md::MatrixDescriptor, key::AbstractString)

Return specific data files (matrix, rhs, solution, or other metadata). The key must be contained in data.metadata or an DataError is thrown. Hint: md.A, md.b, md.x also deliver the matrix, rhs, and solution. So this is needed specifically for other metadata.

MatrixDepot.metastring — Function.

metastring(name, metaname)

In the standard cases convert full meta name to abbreviated form.

given name = "abc", then

"abc.mtx" => "A"

"abc_def.mtx" => "def"

"abc_def.txt" => "def.txt"

"abc-def.mtx" => "abc-def.mtx"

MatrixDepot.metastring_reverse — Function.

metastring_reverse(data::MatrixData, abbreviation::String)

Given a MatrixData object and an abbreviation, return a matchingfull metadata name for the abbreviation. If no full name is found, return the abbreviation unchanged.

MatrixDepot.metasymbols — Function.

metasymbols(md::MatrixDescriptor)

Return a vector of symbols, which point to metadata of the problem. These symbols s may be used to access the objects by getproperty(md, s) or by getindex(md, s). The syntax md.S is preferred, if S is a constant Julia word. In any case md[s] is possible.

Example: md = mdopen("*/bfly"); A = md.A; co = md.coord; txt10 = md["Gname_10.txt"]

MatrixDepot.minij — Function.

MIN[I,J] Matrix

A matrix with (i,j) entry min(i,j). It is a symmetric positive definite matrix. The eigenvalues and eigenvectors are known explicitly. Its inverse is tridiagonal.

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["inverse", "symmetric", "posdef", "eigen"]

References:

J. Fortiana and C. M. Cuadras, A family of matrices, the discretized Brownian bridge, and distance-based regression, Linear Algebra Appl., 264 (1997), 173-188. (For the eigensystem of A.)

MatrixDepot.mksparse — Function.

mksparse(m, n, rowval, colval, nzval)
mksparse!(m, n, rowval, colval, nzval)

Construct a SparseMatrixCSC of dimensions (m,n) from the data given in the three input vectors of equal lengths.

mksparse! destoys the content of colval.

A[rowval[i],colval[i]] == nzval[i] for i ∈ 1:length(nzval). All other entries are zero.

MatrixDepot.mksparse!

MatrixDepot.mm

MatrixDepot.mm_property_name

MatrixDepot.mm_property_type

MatrixDepot.mm_short_name

MatrixDepot.mmread — Function.

mmread(filename|io)

Read Matrixmarket format file (extension .mtx) and return sparse or dense matrix. Symmetric and Hermitian matrices use the corresponding wrapper types. Patterns result in sparse matrics with element type Bool. They may be converted to numerical types by multiplying with a number.

MatrixDepot.mmread_matrix

MatrixDepot.mmreadcomment — Function.

mmreadcomment(filename)

return info comment strings for MatrixMarket format files

MatrixDepot.mmreadheader — Function.

mmreadheader(filename)

Read header information from mtx file.

MatrixDepot.modgroup

MatrixDepot.moler — Function.

Moler Matrix

The Moler matrix is a symmetric positive definite matrix. It has one small eigenvalue.

Input options:

[type,] dim, alpha: dim is the dimension of the matrix, alpha is a scalar.

[type,] dim: alpha = -1.

Groups: ["inverse", "illcond", "symmetric", "posdef"]

References:

J.C. Nash, Compact Numerical Methods for Computers: Linear Algebra and Function Minimisation, second edition, Adam Hilger, Bristol, 1990 (Appendix 1).

MatrixDepot.mtranspose — Function.

mtranspose(A)

Materialized transpose of a matrix

MatrixDepot.neumann — Function.

Neumann Matrix

A singular matrix from the discrete Neumann problem. The matrix is sparse and the null space is formed by a vector of ones

Input options:

[type,] dim: the dimension of the matrix is dim^2.

Groups: ["eigen", "sparse"]

References:

R. J. Plemmons, Regular splittings and the discrete Neumann problem, Numer. Math., 25 (1976), pp. 153-161.

MatrixDepot.newsign

MatrixDepot.oddmagic

MatrixDepot.oscillate — Function.

Oscillating Matrix

A matrix A is called oscillating if A is totally nonnegative and if there exists an integer q > 0 such that A^q is totally positive.

Input options:

[type,] Σ: the singular vaule spectrum of the matrix.

[type,] dim, mode: dim is the dimension of the matrix. mode = 1: geometrically distributed singular values. mode = 2: arithmetrically distributed singular values.

[type,] dim: mode = 1.

Groups: ['symmetric','posdef', 'random', 'eigen']

References:

Per Christian Hansen, Test matrices for regularization methods. SIAM J. SCI. COMPUT Vol 16, No2, pp 506-512 (1995).

MatrixDepot.overview — Function.

overview([db])

return formatted overview about matrices and groups in the collection.

MatrixDepot.parallax — Function.

Stellar Parallax Problem with 26 Fixed, Real Observations

The underlying problem is a Fredholm integral equation of the first kind with kernel

K(s,t) = (1/(σ√(2π)))*exp(-0.5*((s-t)/σ)^2),

with σ = 0.014234, and it is dscretized by means of a Galerkin method with n orthonormal basis functions. The right-hand side b consists of a measured distribution function of stellar parallaxes, and its length is fixed at 26, i.e, the matrix A is 26×n. The exact solution, which represents the true distribution of stellar parallaxes, is unknown.

Input options:

[type,] dim, [matrixonly]: the generated matrix is 26×dim. If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated (matrixonly = true by default).

Groups: ["regprob"]

References:

W. M. Smart, Stellar Dynamics, Cambridge University Press, Cambridge, (1938), p. 30.

MatrixDepot.parse_headerinfo

MatrixDepot.parseint

MatrixDepot.parseloop!

MatrixDepot.parsenext

MatrixDepot.parserr

MatrixDepot.parter — Function.

Parter Matrix

The Parter matrix is a Toeplitz and Cauchy matrix with singular values near π.

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["eigen"]

References:

The MathWorks Newsletter, Volume 1, Issue 1, March 1986, page 2. S. V. Parter, On the distribution of the singular values of Toeplitz matrices, Linear Algebra and Appl., 80 (1986), pp. 115-130.

MatrixDepot.pascal — Function.

Pascal Matrix

The Pascal matrix’s anti-diagonals form the Pascal’s triangle.

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["inverse", "illcond", "symmetric", "posdef", "eigen"]

References:

R. Brawer and M. Pirovino, The linear algebra of the Pascal matrix, Linear Algebra and Appl., 174 (1992), pp. 13-23 (this paper gives a factorization of L = PASCAL(N,1) and a formula for the elements of L^k).

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.4.

MatrixDepot.pei — Function.

Pei Matrix

The Pei matrix is a symmetric matrix with known inversion.

Input options:

[type,] dim, α: dim is the dimension of the matrix. α is a scalar.

[type,] dim: the dimension of the matrix.

Groups: ["inverse", "illcond", "symmetric", "posdef"]

References:

M. L. Pei, A test matrix for inversion procedures, Comm. ACM, 5 (1962), p. 508.

MatrixDepot.phillips — Function.

Phillips's "Famous" Problem

Input options:

[type,] dim, [matrixonly]: the dimenstion of the matrix is dim. If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated (matrixonly = true by default).

Groups: ["regprob"]

References:

D. L. Phillips, A technique for the numerical solution of certain integral equations of the first kind, J. ACM 9 (1962), 84-97.

MatrixDepot.poisson — Function.

Poisson Matrix

A block tridiagonal matrix from Poisson’s equation. This matrix is sparse, symmetric positive definite and has known eigenvalues. The result is of type SparseMatrixCSC.

Input options:

[type,] dim: the dimension of the matirx is dim^2.

Groups: ["inverse", "symmetric", "posdef", "eigen", "sparse"]

References:

G. H. Golub and C. F. Van Loan, Matrix Computations, second edition, Johns Hopkins University Press, Baltimore, Maryland, 1989 (Section 4.5.4).

MatrixDepot.pred

MatrixDepot.prednzdev — Function.

prednzdev(

MatrixDepot.preferred

MatrixDepot.prolate — Function.

Prolate Matrix

A prolate matrix is a symmetirc, ill-conditioned Toeplitz matrix.

Input options:

[type,] dim, w: dim is the dimension of the matrix. w is a real scalar.

[type,] dim: the case when w = 0.25.

References:

J. M. Varah. The Prolate Matrix. Linear Algebra and Appl. 187:267–278, 1993.

MatrixDepot.propline

MatrixDepot.push_hdr!

MatrixDepot.qmult!

MatrixDepot.randcorr — Function.

Random Correlation Matrix

A random correlation matrix is a symmetric positive semidefinite matrix with 1s on the diagonal.

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["symmetric", 'pos-semidef', "random"]

MatrixDepot.rando — Function.

Random Matrix with Element -1, 0, 1

Input options:

[type,] rowdim, coldim, k: row_dim and col_dim are row and column dimensions, k = 1: entries are 0 or 1. k = 2: entries are -1 or 1. k = 3: entries are -1, 0 or 1.

[type,] dim, k: row_dim = col_dim = dim.

[type,] dim: k = 1.

Groups: ["random"]

MatrixDepot.randsvd — Function.

Random Matrix with Pre-assigned Singular Values

Input options:

[type,] rowdim, coldim, kappa, mode: row_dim and col_dim are the row and column dimensions. kappa is the condition number of the matrix. mode = 1: one large singular value. mode = 2: one small singular value. mode = 3: geometrically distributed singular values. mode = 4: arithmetrically distributed singular values. mode = 5: random singular values with unif. dist. logarithm.

[type,] dim, kappa, mode: row_dim = col_dim = dim.

[type,] dim, kappa: mode = 3.

[type,] dim: kappa = sqrt(1/eps()), mode = 3.

Groups: ["illcond", "random"]

References:

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.3.

MatrixDepot.remotetype

MatrixDepot.resall!

MatrixDepot.rohess — Function.

Random Orthogonal Upper Hessenberg Matrix

The matrix is constructed via a product of Givens rotations.

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["random"]

References:

W. B. Gragg, The QR algorithm for unitary Hessenberg matrices, J. Comp. Appl. Math., 16 (1986), pp. 1-8.

MatrixDepot.rosser — Function.

Rosser Matrix

The Rosser matrix’s eigenvalues are very close together so it is a challenging matrix for many eigenvalue algorithms.

Input options:

[type,] dim, a, b: dim is the dimension of the matrix. dim must be a power of 2. a and b are scalars. For dim = 8, a = 2 and b = 1, the generated matrix is the test matrix used by Rosser.

[type,] dim: a = rand(1:5), b = rand(1:5).

Groups: ["eigen", "illcond", "random"]

References:

J. B. Rosser, C. Lanczos, M. R. Hestenes, W. Karush, Separation of close eigenvalues of a real symmetric matrix, Journal of Research of the National Bureau of Standards, v(47) (1951)

MatrixDepot.round_matlab

MatrixDepot.sampling — Function.

Matrix with Application in Sampling Theory

A nonsymmetric matrix with eigenvalues 0, 1, 2, ... n-1.

Input options:

[type,] vec: vec is a vector with no repeated elements.

[type,] dim: dim is the dimension of the matrix. vec = [1:dim;]/dim.

Groups: ["eigen"]

References:

L. Bondesson and I. Traat, A nonsymmetric matrix with integer eigenvalues, linear and multilinear algebra, 55(3) (2007), pp. 239-247

MatrixDepot.shaw — Function.

One-Dimensional Image Restoration Model

This test problem uses a first-kind Fredholm integral equation to model a one-dimentional image restoration situation.

Input options:

[type,] dim, [matrixonly]: the dimesion of the matrix dim must be even. If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated (matrixonly = true by default).

Groups: ["regprob"]

References:

C. B. Shaw, Jr., Improvements of the resolution of an instrument by numerical solution of an integral equation. J. Math. Ana. Appl. 37 (1972), 83-112.

MatrixDepot.shell_to_regex — Function.

shell_to_regex

return a regular expression if shell pattern characters "*?]" are contained in string, otherwise return string. If no '/' is contained in p and p is not "", insert `"(./)?"` in regular expression.

MatrixDepot.shortcuts

MatrixDepot.singlist!

MatrixDepot.smallworld — Function.

Small World Network

Generate an adjacency matrix for a small world network. We model it by adding shortcuts to a kth nearest neighbour ring network (nodes i and j are connected iff |i -j| <= k or |n - |i -j|| <= k.) with n nodes.

Input options:

[type,] n, k, p: the dimension of the matrix is n. The number of nearest-neighbours to connect is k. The probability of adding a shortcut in a given row is p.

[type,] n: k = 2 and p = 0.1.

References:

D. J. Watts and S. H. Strogatz, Collective Dynamics of Small World Networks, Nature 393 (1998), pp. 440-442.

MatrixDepot.sort_by_type

MatrixDepot.sp

MatrixDepot.specialsort

MatrixDepot.spikes — Function.

Test Problem with "Spike" Solution

Artifically generated discrete ill-posed problem.

Input options:

[type,] dim, tmax, [matrixonly]: dim is the dimension of the matrix. `tmaxcontrols the length of the pulse train. Ifmatrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated (matrixonly = true` by default). The solution x consists a unit step at t = .5 and a pulse train of spike of decreasing magnitude at t = .5, 1.5, 2.5, ....

[type,] dim, [matrixonly]: t_max = 5.

Groups: ["regprob"]

MatrixDepot.sumwidth

MatrixDepot.tab

MatrixDepot.toeplitz — Function.

Toeplitz Matrix

A Toeplitz matrix is a matrix in which each descending diagonal from left to right is constant.

Input options:

[type,] vc, vr: vc and vr are the first column and row of the matrix.

[type,] v: symmatric case, i.e., vc = vr = v.

[type,] dim: dim is the dimension of the matrix. v = [1:dim;] is the first row and column vector.

MatrixDepot.toggle_remote

MatrixDepot.tomo

MatrixDepot.tridiag — Function.

Tridiagonal Matrix

Construct a tridigonal matrix of type Tridiagonal.

Input options:

[type,] v1, v2, v3: v1 and v3 are vectors of subdiagonal and superdiagonal elements, respectively, and v2 is a vector of diagonal elements.

[type,] dim, x, y, z: dim is the dimension of the matrix, x, y, z are scalars. x and z are the subdiagonal and superdiagonal elements, respectively, and y is the diagonal elements.

[type,] dim: x = -1, y = 2, z = -1. This matrix is also known as the second difference matrix.

Groups: ["inverse", "illcond", "posdef", "eigen"]

References:

J. Todd, Basic Numerical Mathematics, Vol. 2: Numerical Algebra, Birkhauser, Basel, and Academic Press, New York, 1977, p. 155.

MatrixDepot.triw — Function.

Triw Matrix

Upper triangular matrices discussed by Wilkinson and others.

Input options:

[type,] rowdim, coldim, α, k: row_dim and col_dim are row and column dimension of the matrix. α is a scalar representing the entries on the superdiagonals. k is the number of superdiagonals.

[type,] dim: the dimension of the matrix.

Groups: ["inverse", "illcond"]

References:

G. H. Golub and J. H. Wilkinson, Ill-conditioned eigensystems and the computation of the Jordan canonical form, SIAM Review, 18(4), 1976, pp. 578-6

MatrixDepot.uf_remote

MatrixDepot.update — Function.

MatrixDepot.update()

update database index from the websites

MatrixDepot.ursell — Function.

Integral Equation with No square Integrable Solution

Discretization of a first kind Fredholm integral equation with kernel K and right-hand side g given by K(s,t) = 1/(s+t+1), g(s) = 1, where both integration intervals are [0, 1]. The matrix A is a Hankel matrix.

Input options:

[type,] dim, [matrixonly]: dim is the dimension of the matrix. If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will also be generated (matrixonly = true by default).

Groups: ["regprob"]

References:

F. Ursell, Introduction to the theory of linear integral equations., Chapter 1 in L. M. Delves & J. Walsh, Numerical Solution of Integral Equations, Clarendon Press, 1974.

MatrixDepot.user

MatrixDepot.usermatrixclass

MatrixDepot.vand — Function.

Vandermonde Matrix

The inverse and determinat are known explicity.

Input options:

[type,] vec, coldim: vec is a vector, `coldim` is the number of columns.

[type,] vec: col_dim = length(vec)

[type,] dim: vec = [1:dim;]

Groups: ["inverse", "illcond"]

References:

N. J. Higham, Stability analysis of algorithms for solving confluent Vandermonde-like systems, SIAM J. Matrix Anal. Appl., 11 (1990), pp. 23-41.

MatrixDepot.vecvec

MatrixDepot.verify_loaded

MatrixDepot.verify_loadinfo

MatrixDepot.wathen — Function.

Wathen Matrix

Wathen Matrix is a sparse, symmetric positive, random matrix arose from the finite element method. The generated matrix is the consistent mass matrix for a regular nx-by-ny grid of 8-nodes.

Input options:

[type,] nx, ny: the dimension of the matrix is equal to 3 * nx * ny + 2 * nx * ny + 1.

[type,] n: nx = ny = n.

Groups: ["symmetric", "posdef", "eigen", "random", "sparse"]

References:

A. J. Wathen, Realistic eigenvalue bounds for the Galerkin mass matrix, IMA J. Numer. Anal., 7 (1987), pp. 449-457.

MatrixDepot.wilkinson — Function.

Wilkinson Matrix

The Wilkinson matrix is a symmetric tridiagonal matrix with pairs of nearly equal eigenvalues. The most frequently used case is matrixdepot("wilkinson", 21). The result is of type Tridiagonal.

Input options:

[type,] dim: the dimension of the matrix.

Groups: ["symmetric", "eigen"]

References:

J. H. Wilkinson, Error analysis of direct methods of matrix inversion, J. Assoc. Comput. Mach., 8 (1961), pp. 281-330.

MatrixDepot.wing — Function.

A Problem with a Discontinuous Solution

Input options:

[type,] dim, t1, t2, [matrixonly]: the dimension of matrix is dim. t1 and t2 are two real scalars such that 0 < t1 < t2 < 1. If matrixonly = false, the matrix A and vectors b and x in the linear system Ax = b will be generated (matrixonly = true by default).

[type,] n, [matrixonly]: t1 = 1/3 and t2 = 2/3.

Groups: ["regprob"]

References:

G. M. Wing, A Primer on Integral Equations of the First Kind, Society for Industrial and Applied Mathematics, 1991, p. 109.

MatrixDepot.wordlist — Function.

wordlist(string)

Separate words is string by spaces and delimiters. Return list of unique words, which can be used as keywords.

MatrixDepot.wrap