Docstrings

Dagger.@dbg — Macro.

Run a block of code only if DEBUG is true

Dagger.@logmsg

Dagger.@unimplemented — Macro.

@unimplemented fname(<args...>)

While it is nice to define generic function ad-hoc, it can sometimes get confusing to figure out which method is missing. @interface creates a function which errors out pointing which method is missing.

Dagger.AbstractPart

Dagger.AllocateArray

Dagger.ArrayDomain

Dagger.ArrayOp

Dagger.B

Dagger.BCast

Dagger.BlockPartition

Dagger.BlockedDomains

Dagger.Blocks

Dagger.Bytes

Dagger.CAT

Dagger.Cat

Dagger.Chunk — Type.

A chunk with some data

Dagger.Computation

Dagger.ComputedArray

Dagger.Concat

Dagger.Context — Type.

A context represents a set of processors to use for a papply operation.

Dagger.DArray — Type.

DArray{T,N,F}(domain, subdomains, chunks, concat)

An N-dimensional distributed array of element type T.

domain: the whole ArrayDomain of the array

subdomains: a DomainBlocks of the same dimensions as the array

chunks: an array of chunks of dimension N

concat: a function of type F. concat(x, y; dims=d) takes two chunks x and y and concatenates them along dimension d. cat is used by default.

Dagger.Dagger

Dagger.DaggerBroadcastStyle

Dagger.Distribute

Dagger.DomainBlocks

Dagger.ENABLE_DEBUG

Dagger.Event

Dagger.FileReader

Dagger.FilterLog

Dagger.GB

Dagger.GetIndex

Dagger.GetIndexScalar

Dagger.LocalEventLog — Type.

represents a process local events array.

A context with log_sink set to LocalEventLog() will cause events to be recorded into the

Dagger.MB

Dagger.Map

Dagger.MapChunk

Dagger.MatMul

Dagger.NoOpLog — Type.

Various means of writing an event to something.

Dagger.OSProc — Type.

OS process - contains pid returned by addprocs

Dagger.PARTSPEC

Dagger.PLUGINS

Dagger.PLUGIN_CONFIGS

Dagger.Part

Dagger.Processor

Dagger.ProfilerResult

Dagger.PromotePartition — Type.

This is a way of suggesting that stage should call stage_operand with the operation and other arguments

Dagger.ReduceBlock

Dagger.Reducedim

Dagger.Save

Dagger.Scale

Dagger.Sch

Dagger.SetIndex

Dagger.State — Type.

Overall state used during visualization

Dagger.TB

Dagger.Thunk

Dagger.Timespan — Type.

identifies

space (category, id) time (timeline, start, finish)

also tracks gc_num during this and profiling samples.

Dagger.Timestamp

Dagger.Transpose

Dagger.UnitDomain — Type.

Default domain – has no information about the value

Dagger._ctranspose

Dagger._cumlength

Dagger._cumsum

Dagger._dep_message_AbstractPart

Dagger._dep_message_BlockedDomains

Dagger._dep_message_Cat

Dagger._dep_message_ComputedArray

Dagger._dep_message_Part

Dagger._getindex

Dagger._intersect

Dagger._local_event_log

Dagger._merge_sorted

Dagger._mul

Dagger._part_labels

Dagger._promote_array

Dagger._scale

Dagger._stage_cache

Dagger._sum

Dagger.add_gc_diff

Dagger.add_span — Function.

Add a Timespan to a given State under tl (timeline) and category.

Dagger.affinity

Dagger.aggregate_events

Dagger.alignfirst — Function.

alignfirst(a)

Make a subdomain a standalone domain. For example,

alignfirst(ArrayDomain(11:25, 21:100))
# => ArrayDomain((1:15), (1:80))

Dagger.allslices

Dagger.argname

Dagger.arrayorvcat

Dagger.batchedsplitmerge

Dagger.build_timespans

Dagger.cache_result!

Dagger.cached_stage — Function.

A memoized version of stage. It is important that the tasks generated for the same DArray have the same identity, for example:

A = rand(Blocks(100,100), Float64, 1000, 1000)
compute(A+A')

must not result in computation of A twice.

Dagger.chunks

Dagger.chunktype

Dagger.cleanup

Dagger.clear_local_event_log

Dagger.collect_merge

Dagger.compute — Function.

Compute a Thunk - creates the DAG, assigns ranks to nodes for tie breaking and runs the scheduler.

A DArray object may contain a thunk in it, in which case we first turn it into a Thunk object and then compute it.

Dagger.cumulative_domains

Dagger.debug_compute

Dagger.default_ord

Dagger.delayed

Dagger.delayedmap

Dagger.dependents — Function.

find the set of direct dependents for each task

Dagger.distribute

Dagger.domain — Function.

domain(x::T)

Returns metadata about x. This metadata will be in the domain field of a Chunk object when an object of type T is created as the result of evaluating a Thunk.

Dagger.domainchunks

Dagger.dsort_chunks — Function.

dsort_chunks(cs, [nchunks, nsamples]; options...)

Sort contents of chunks (cs) and return a new set of chunks such that the chunks when concatenated return a sorted collection. Each chunk in turn is sorted.

Args:

nchunks – the number of chunks to produce, regardless of how many chunks were given as input

nsamples – the number of elements to sample from each chunk to guess the splitters (nchunks-1 splitters) each chunk will be delimited by the splitter.

merge – a function to merge two sorted collections.

sub – a function to get a subset of the collection takes (collection, range) (defaults to getindex)

order – Base.Sort.Ordering to be used for sorting

batchsize – number of chunks to split and merge at a time (e.g. if there are 128 input chunks and 128 output chunks, and batchsize is 8, then we first sort among batches of 8 chunks – giving 16 batches. Then we sort among the first chunk of the first 8 batches (all elements less than the first splitter), then go on to the first 8 chunks of the second 8 batches, and so on...

chunks_presorted – is each chunk in the input already sorted?

sortandsample – a function to sort a chunk, then sample N elements to infer the splitters. It takes 3 arguments: (collection, N, presorted). presorted is a boolean which is true if the chunk is already sorted.

Returns a tuple of (chunk, samples) where chunk is the Dagger.Chunk object. chunk can be nothing if no change to the initial array was made (e.g. it was already sorted)

affinities – a list of processes where the output chunks should go. If the length is not equal to nchunks then affinities array is cycled through.

Dagger.empty_prof

Dagger.eval

Dagger.fnbody

Dagger.free!

Dagger.free_chunks

Dagger.gather

Dagger.get_logs! — Function.

Get the logs from each process, clear it too

Dagger.get_type

Dagger.getmedians

Dagger.getsub

Dagger.getvarname

Dagger.group_indices

Dagger.include

Dagger.indexes

Dagger.inputs

Dagger.isasync

Dagger.istask

Dagger.kB

Dagger.load — Function.

load(ctx, file_path)

Load an Union{Chunk, Thunk} from a file.

load(ctx, ::Type{Chunk}, fpath, io)

Load a Chunk object from a file, the file path is required for creating a FileReader object

Dagger.lookup_parts

Dagger.make_timespan — Function.

create a timespan given the strt and finish events

Dagger.mapchunk

Dagger.mappart

Dagger.mapreducebykey_seq

Dagger.merge_cumsums

Dagger.merge_reducebykey

Dagger.merge_sorted

Dagger.mix_samples

Dagger.mul_size

Dagger.next_id

Dagger.next_state — Function.

When building state for real-time visualization, use next_state to progress gantt state.

Dagger.node_id

Dagger.node_label

Dagger.noffspring — Function.

recursively find the number of taks dependent on each task in the DAG. Input: dependents dict

Dagger.order — Function.

Given a root node of the DAG, calculates a total order for tie-braking

Root node gets score 1,

rest of the nodes are explored in DFS fashion but chunks of each node are explored in order of noffspring, i.e. total number of tasks depending on the result of the said node.

Args: - node: root node - ndeps: result of noffspring

Dagger.part

Dagger.partition

Dagger.partition_sparse

Dagger.parts

Dagger.parttype

Dagger.persist!

Dagger.project

Dagger.promote_distribution

Dagger.propagate_affinity!

Dagger.pushkey

Dagger.raise_event

Dagger.recursive_splitters — Function.

split the splitters themselves into batches.

Args:

ord – Sorting.Ordering object

splitters – the nchunks-1 splitters

batchsize – batch size

Returns: A Tuple{Vector, Vector{Vector}} – the coarse splitters which create batchsize splits, finer splitters within those batches which create a total of nchunks splits.

julia> Dagger.recursive_splitters(Dagger.default_ord,
            [10,20,30,40,50,60], 5,3)
([30], Any[[10, 20], [40, 50, 60]])

The first value [30] represents a coarse split that cuts the dataset from -Inf-30, and 30-Inf. Each part is further recursively split using the next set of splitters

Dagger.reduce_async

Dagger.reduceblock

Dagger.reduceblock_async

Dagger.reducebykey

Dagger.reducebykey_seq

Dagger.reduced_dims

Dagger.save — Function.

save(io::IO, val)

Save a value into the IO buffer. In the case of arrays and sparse matrices, this will save it in a memory-mappable way.

load(io::IO, t::Type, domain) will load the object given its domain

save(ctx, chunk::Union{Chunk, Thunk}, file_path::AbsractString)

Save a chunk to a file at file_path.

special case distmem writing - write to disk on the process with the chunk.

Dagger.savechunk

Dagger.scale

Dagger.setindex

Dagger.shouldpersist

Dagger.show_ast

Dagger.show_plan

Dagger.show_statement

Dagger.showfn

Dagger.showloc

Dagger.sortandsample_array

Dagger.split_range — Function.

Utility function to divide the range range into n chunks

Dagger.split_range_interval — Function.

split_range_interval(range, n)

split a range into pieces each of length n or lesser

Dagger.splitchunk

Dagger.splitmerge

Dagger.stage

Dagger.stage_operand

Dagger.stage_operands — Function.

an operand which should be distributed as per convenience

Dagger.summarize_events

Dagger.thunkize — Function.

If a DArray tree has a Thunk in it, make the whole thing a big thunk

Dagger.timespan_end

Dagger.timespan_start

Dagger.tochunk — Function.

Create a chunk from a sequential object.

Dagger.transpose_vecvec

Dagger.treereduce — Function.

Tree reduce

Dagger.treereduce_nd

Dagger.treereducedim

Dagger.unrelease

Dagger.use_shared_array

Dagger.write_dag

Dagger.write_event — Function.

Write a log event