Docstrings

Wavelets.BiggestTH

Wavelets.DNFT

Wavelets.DiscreteWavelet

Wavelets.Entropy

Wavelets.FilterWavelet

Wavelets.WT.GLS — Type.

Wavelet type for discrete general (bi)orthogonal transforms by using a lifting scheme.

See also: OrthoFilter, wavelet

Wavelets.HardTH

Wavelets.LSWavelet

Wavelets.LogEnergyEntropy

Wavelets.NegTH

Wavelets.WT.OrthoFilter — Type.

Wavelet type for discrete orthogonal transforms by filtering.

See also: GLS, wavelet

Wavelets.Plot

Wavelets.PosTH

Wavelets.SemiSoftTH

Wavelets.ShannonEntropy

Wavelets.SoftTH

Wavelets.SteinTH

Wavelets.Threshold

Wavelets.Transforms

Wavelets.Util

Wavelets.VisuShrink

Wavelets.WT

Wavelets.Wavelets

Wavelets.bestbasistree

Wavelets.coefentropy

Wavelets.denoise

Wavelets.detailindex

Wavelets.detailn

Wavelets.detailrange

Wavelets.downsample

Wavelets.Transforms.dwt — Function.

dwt(x, wt[, L=maxtransformlevels(x)])

Perform a discrete wavelet transform of the array x. The wavelet type wt determines the transform type (filter or lifting) and the wavelet class, see wavelet.

The number of transformation levels L can be any non-negative integer such that the size of x is divisible by L. Performs the identity transformation if L==0.

Examples

dwt(x, wavelet(WT.coif6))

See also: idwt, dwt!, wavelet

Wavelets.Transforms.dwt! — Function.

dwt!(y, x, wt::OrthoFilter[, L=maxtransformlevels(x)])

dwt!(y, wt::GLS[, L=maxtransformlevels(x)])

Same as dwt but without array allocation. Perform "out of place" transform with a filter, or a inplace transform with a lifting scheme. The difference between the filter and lifting methods is due to the structure of the transform algorithms.

See also: idwt!

Wavelets.dyadicdetailindex

Wavelets.dyadicdetailn

Wavelets.dyadicdetailrange

Wavelets.dyadiclevel2tl

Wavelets.dyadicscalingrange

Wavelets.eval

Wavelets.Transforms.idwt — Function.

idwt(x, wt[, L=maxtransformlevels(x)])

Perform an inverse discrete wavelet transform of the array x, the inverse of dwt(x, wt, L).

See also: dwt, idwt!

Wavelets.Transforms.idwt! — Function.

idwt!(y, x, wt::OrthoFilter[, L=maxtransformlevels(x)])

idwt!(y, wt::GLS[, L=maxtransformlevels(x)])

The inverse of dwt!.

See also: dwt!

Wavelets.Transforms.imodwt — Function.

Perform an inverse maximal overlap discrete wavelet transform (MODWT) of xw, the inverse of modwt(x, wt, L).

Wavelets.include

Wavelets.iscube

Wavelets.isdyadic

Wavelets.isvalidtree

Wavelets.Transforms.iwpt — Function.

iwpt

Perform an inverse discrete wavelet packet transform of the array x. See also: idwt, wavelet

Wavelets.Transforms.iwpt! — Function.

iwpt!

Same as iwpt but without array allocation.

Wavelets.maketree

Wavelets.makewavelet

Wavelets.matchingpursuit

Wavelets.maxdyadiclevel

Wavelets.maxmodwttransformlevels

Wavelets.maxtransformlevels

Wavelets.mirror

Wavelets.Transforms.modwt — Function.

Perform a maximal overlap discrete wavelet transform (MODWT) of the array x. The wavelet type wt determines the transform type and the wavelet class, see wavelet. (Note that the wavelet filter coefficients are scaled by 1/√2 for the MODWT so that the transform maintains unit energy).

The number of transform levels L can be any number <= maxmodwttransformlevels(x) (the default value).

Returns an n × L+1 matrix (where n is the length of x) with the wavelet coefficients for level j in column j. The scaling coefficients are in the last (L+1th) column.

Wavelets.ndyadicscales

Wavelets.noisest

Wavelets.stridedcopy!

Wavelets.sufficientpoweroftwo

Wavelets.testfunction

Wavelets.threshold

Wavelets.threshold!

Wavelets.tl2dyadiclevel

Wavelets.upsample

Wavelets.WT.wavelet — Function.

wavelet(c[, t=WT.Filter][, boundary=WT.Periodic])

Construct wavelet type where c is a wavelet class, t is the transformation type (WT.Filter or WT.Lifting), and boundary is the type of boundary treatment.

Examples

wavelet(WT.coif6)
wavelet(WT.db1, WT.Lifting)

See also: WT.WaveletClass

Wavelets.wcount

Wavelets.wplotdots

Wavelets.wplotim

Wavelets.Transforms.wpt — Function.

wpt

Perform a discrete wavelet packet transform of the array x. See also: dwt, wavelet

Wavelets.Transforms.wpt! — Function.

wpt!

Same as wpt but without array allocation.