Docstrings

BigArrays.AbstractBigArray

BigArrays.BackendBase

BigArrays.BigArray — Type.

BigArray

currently, assume that the array dimension (x,y,z,...) is >= 3 all the manipulation effects in the x,y,z dimension

BigArrays.BigArrays

BigArrays.BinDicts

BigArrays.CHUNK_CHANNEL_SIZE

BigArrays.ChunkIterators

BigArrays.Codings

BigArrays.DEFAULT_FILL_MISSING

BigArrays.DEFAULT_MODE

BigArrays.GSDicts

BigArrays.GZIP_MAGIC_NUMBER

BigArrays.Indexes

BigArrays.Infos

BigArrays.S3Dicts

BigArrays.TASK_NUM

BigArrays.WORKER_POOL

BigArrays.adjust_volume_boundary — Function.

adjust the global and buffer range according to total volume size. shrink the range stop if the ranges passes the volume boundary.

BigArrays.eval

BigArrays.get_chunk_size

BigArrays.get_mode

BigArrays.get_num_chunks — Function.

get_num_chunks(ba::BigArray, idxes::Union{UnitRange,Int}...)

get number of chunks needed to do cutout from this range

BigArrays.getindex_multiprocesses

BigArrays.getindex_multiprocesses_worker

BigArrays.getindex_multithreads

BigArrays.getindex_multithreads_worker!

BigArrays.getindex_sequential — Function.

get_index_sequential(ba::BigArray, idxes::Union{UnitRange, Int}...)

sequential implementation for debuging

BigArrays.getindex_sharedarray

BigArrays.getindex_sharedarray_worker!

BigArrays.include

BigArrays.list_missing_chunks — Function.

list_missing_chunks(ba::BigArray, idxes::Union{UnitRange, Int}...)

list the non-existing keys in the index range if the returned list is empty, then all the chunks exist in the storage backend.

BigArrays.set_chunk_size

BigArrays.set_mode

BigArrays.setindex_multiprocesses! — Function.

put array in RAM to a BigArray

this version uses channel to control the number of asynchronized request

BigArrays.setindex_multiprocesses_worker

BigArrays.setindex_multithreads! — Function.

put array in RAM to a BigArray backend

this version uses channel to control the number of asynchronized request

BigArrays.setindex_multithreads_worker

BigArrays.setindex_sequential! — Function.

sequential function, good for debuging

BigArrays.setindex_sharedarray! — Function.

put array in RAM to a BigArray

this version uses channel to control the number of asynchronized request

BigArrays.setindex_sharedarray_worker