Docstrings

ZipFile.Deflate — Constant.

Deflate compression method

ZipFile.Dir — Type.

Reader represents a ZIP file open for reading.

Reader(io::IO)
Reader(filename::AbstractString)

Read a ZIP file from io or the file named filename.

ZipFile.File

ZipFile.ReadableFile

ZipFile.Reader — Type.

Reader represents a ZIP file open for reading.

Reader(io::IO)
Reader(filename::AbstractString)

Read a ZIP file from io or the file named filename.

ZipFile.Store — Constant.

Compression method that does no compression

ZipFile.WritableDir — Type.

Reader represents a ZIP file open for writing.

Writer(io::IO)
Writer(filename::AbstractString)

Create a new ZIP file that will be written to io or the file named filename.

ZipFile.WritableFile

ZipFile.Writer — Type.

Reader represents a ZIP file open for writing.

Writer(io::IO)
Writer(filename::AbstractString)

Create a new ZIP file that will be written to io or the file named filename.

ZipFile.ZipFile — Module.

A Julia package for reading/writing ZIP archive files

This package provides support for reading and writing ZIP archives in Julia. Install it via the Julia package manager using $Pkg.add("ZipFile")$.

The ZIP file format is described in http://www.pkware.com/documents/casestudies/APPNOTE.TXT

Example

The example below writes a new ZIP file and then reads back the contents.

julia> using ZipFile
julia> w = ZipFile.Writer("/tmp/example.zip");
julia> f = ZipFile.addfile(w, "hello.txt");
julia> write(f, "hello world!
");
julia> f = ZipFile.addfile(w, "julia.txt", method=ZipFile.Deflate);
julia> write(f, "Julia
"^5);
julia> close(w)
julia> r = ZipFile.Reader("/tmp/example.zip");
julia> for f in r.files
          println("Filename: $(f.name)")
          write(stdout, readstring(f));
       end
julia> close(r)
Filename: hello.txt
hello world!
Filename: julia.txt
Julia
Julia
Julia
Julia
Julia

ZipFile.Zlib

ZipFile._CentralDirSig

ZipFile._EndCentralDirSig

ZipFile._LocalFileHdrSig

ZipFile._Method2Str

ZipFile._ZipVersion

ZipFile._find_diroffset

ZipFile._find_enddiroffset

ZipFile._getfiles

ZipFile._msdostime

ZipFile._mtime

ZipFile._writele

ZipFile.addfile — Function.

addfile(w::Writer, name::AbstractString; method::Integer=Store, mtime::Float64=-1.0)

Add a new file named name into the ZIP file writer w, and return the WritableFile for the new file. We don't allow concurrrent writes, thus the file previously added using this function will be closed.

Method specifies the compression method that will be used (Store for uncompressed or Deflate for compressed).

Mtime is the modification time of the file.

ZipFile.close

ZipFile.eof

ZipFile.eval

ZipFile.getindex_u32le — Function.

Load a little endian UInt32 from a UInt8 vector b starting from index i

ZipFile.include

ZipFile.mtime

ZipFile.open

ZipFile.position

ZipFile.read

ZipFile.readle

ZipFile.show

ZipFile.utf8_validate

ZipFile.write