Docstrings

SimpleTraits.@check_fast_traitdispatch — Macro.

check_fast_traitdispatch(Tr, Args=(Int,), nlines=6, verbose=false)

Macro to check whether a trait-dispatch is fast (i.e. as fast as an ordinary function call) or whether dispatch is slow (dynamic). Only works with single parameters traits (so far).

Optional arguments are:

Type parameter to the trait (default Int)

Verbosity (default false)

Example:

@check_fast_traitdispatch IsBits
@check_fast_traitdispatch IsBits String true

NOTE: This only works when code-coverage is disabled! Thus, do not use this macro in tests (or disable coverage=true in your .travis.yml script), or it will error.

NOTE: This does not seem to work all the time. At least in Julia 0.7 this test is negative for IsIndexLinear, however, it's fast.

TODO: This is rather ugly. Ideally this would be a function but I ran into problems, see source code. Also the macro is ugly. PRs welcome...

SimpleTraits.@traitdef — Macro.

Used to define a trait. Traits, like types, are camel cased. I suggest to start them with a verb, e.g. IsImmutable, to distinguish them from actual types, which are usually nouns.

Traits need to have one or more (type-)parameters to specify the type to which the trait is applied. For instance IsImmutable{Int} signifies that Int is part of IsImmutable (although whether that is true needs to be checked with the istrait function). Most traits will be one-parameter traits, however, several parameters are useful when there is a "contract" between several types.

Examples:

@traitdef IsFast{X}
@traitdef IsSlow{X,Y}

SimpleTraits.@traitfn — Macro.

Defines a function dispatching on a trait. Examples:

@traitfn f(x::X,y) where {X;  Tr1{X}} = ...
@traitfn f(x::X,y) where {X; !Tr1{X}} = ...

@traitfn f(x::X,y::Y) where {X,Y;  Tr2{X,Y}} = ...
@traitfn f(x::X,y::Y) where {X,Y; !Tr2{X,Y}} = ...

Note that the second example is just syntax sugar for @traitfn f(x::X,y::Y) where {X,Y; Not{Tr1{X,Y}}} = ....

SimpleTraits.@traitimpl — Macro.

Used to add a type or type-tuple to a trait. By default a type does not belong to a trait.

Example:

@traitdef IsFast{X}
@traitimpl IsFast{Array{Int,1}}

Often a trait is dependent on some check-function returning true or false. This can be done with:

@traitimpl IsFast{T} <- isfast(T)

where isfast is that check-function.

Note that traits implemented with the former of above methods will override an implementation with the latter method. Thus it can be used to define exceptions to the rule.

SimpleTraits.BaseTraits

SimpleTraits.GenerateTypeVars

SimpleTraits.Not — Type.

The set of all types not belonging to a trait is encoded by wrapping it with Not{}, e.g. Not{Tr1{X,Y}}

SimpleTraits.SimpleTraits

SimpleTraits.Trait — Type.

All Traits are subtypes of abstract type Trait.

A concrete Trait will look like:

struct Tr1{X,Y} <: Trait end

where X and Y are the types involved in the trait.

(SUPER is not used here but in Traits.jl, thus retained for possible future compatibility.)

SimpleTraits.curmod

SimpleTraits.dispatch_cache

SimpleTraits.eval

SimpleTraits.findline

SimpleTraits.include

SimpleTraits.insertdummy

SimpleTraits.isnegated

SimpleTraits.istrait — Function.

This function checks whether a trait is fulfilled by a specific set of types.

istrait(Tr1{Int,Float64}) => return true or false

SimpleTraits.llvm_lines — Function.

Returns number of llvm-IR lines for a call of function fn with argument types args

SimpleTraits.stripNot

SimpleTraits.strip_kw

SimpleTraits.strip_tpara

SimpleTraits.trait — Function.

A trait is defined as full-filled if this function is the identity function for that trait. Otherwise it returns the trait wrapped in Not.

Example:

trait(IsBits{Int}) # returns IsBits{Int}
trait(IsBits{Array}) # returns Not{IsBits{Array}}

Usually this function is defined when using the @traitimpl macro.

However, instead of using @traitimpl one can define a method for trait to implement a trait, see the README.

SimpleTraits.traitfn