Docstrings

StatsFuns.RFunctions

StatsFuns.StatsFuns

StatsFuns._log1pmx_ker

StatsFuns._norminvlogcdf_impl

StatsFuns._qnorm_ker1

StatsFuns._qnorm_ker2

StatsFuns.betaccdf

StatsFuns.betacdf

StatsFuns.betainvccdf

StatsFuns.betainvcdf

StatsFuns.betainvlogccdf

StatsFuns.betainvlogcdf

StatsFuns.betalogccdf

StatsFuns.betalogcdf

StatsFuns.betalogpdf

StatsFuns.betapdf

StatsFuns.binomccdf

StatsFuns.binomcdf

StatsFuns.binominvccdf

StatsFuns.binominvcdf

StatsFuns.binominvlogccdf

StatsFuns.binominvlogcdf

StatsFuns.binomlogccdf

StatsFuns.binomlogcdf

StatsFuns.binomlogpdf

StatsFuns.binompdf

StatsFuns.chisqccdf

StatsFuns.chisqcdf

StatsFuns.chisqinvccdf

StatsFuns.chisqinvcdf

StatsFuns.chisqinvlogccdf

StatsFuns.chisqinvlogcdf

StatsFuns.chisqlogccdf

StatsFuns.chisqlogcdf

StatsFuns.chisqlogpdf

StatsFuns.chisqpdf

StatsFuns.eval

StatsFuns.fdistccdf

StatsFuns.fdistcdf

StatsFuns.fdistinvccdf

StatsFuns.fdistinvcdf

StatsFuns.fdistinvlogccdf

StatsFuns.fdistinvlogcdf

StatsFuns.fdistlogccdf

StatsFuns.fdistlogcdf

StatsFuns.fdistlogpdf

StatsFuns.fdistpdf

StatsFuns.fourinvπ

StatsFuns.fourπ

StatsFuns.gammaccdf

StatsFuns.gammacdf

StatsFuns.gammainvccdf

StatsFuns.gammainvcdf

StatsFuns.gammainvlogccdf

StatsFuns.gammainvlogcdf

StatsFuns.gammalogccdf

StatsFuns.gammalogcdf

StatsFuns.gammalogpdf

StatsFuns.gammapdf

StatsFuns.halfπ

StatsFuns.hyperccdf

StatsFuns.hypercdf

StatsFuns.hyperinvccdf

StatsFuns.hyperinvcdf

StatsFuns.hyperinvlogccdf

StatsFuns.hyperinvlogcdf

StatsFuns.hyperlogccdf

StatsFuns.hyperlogcdf

StatsFuns.hyperlogpdf

StatsFuns.hyperpdf

StatsFuns.include

StatsFuns.inv2π

StatsFuns.inv4π

StatsFuns.invsoftplus — Function.

logexpm1(x::Real)

Return log(exp(x) - 1) or the "invsoftplus" function. It is the inverse of log1pexp (aka "softplus").

StatsFuns.invsqrt2

StatsFuns.invsqrt2π

StatsFuns.invπ

StatsFuns.log1mexp — Function.

log1mexp(x::Real)

Return log(1 - exp(x))

See:

Martin Maechler (2012) "Accurately Computing log(1 − exp(− |a|))", http://cran.r-project.org/web/packages/Rmpfr/vignettes/log1mexp-note.pdf

Note: different than Maechler (2012), no negation inside parentheses

StatsFuns.log1pexp — Function.

log1pexp(x::Real)

Return log(1+exp(x)) evaluated carefully for largish x.

This is also called the "softplus" transformation, being a smooth approximation to max(0,x). Its inverse is logexpm1.

StatsFuns.log1pmx — Function.

log1pmx(x::Float64)

Return log(1 + x) - x

Use naive calculation or range reduction outside kernel range. Accurate ~2ulps for all x.

StatsFuns.log1psq — Function.

log1psq(x::Real)

Return log(1+x^2) evaluated carefully for abs(x) very small or very large.

StatsFuns.log2mexp — Function.

log2mexp(x::Real)

Return log(2 - exp(x)) evaluated as log1p(-expm1(x))

StatsFuns.log2π

StatsFuns.log4π

StatsFuns.logaddexp — Function.

logaddexp(x::Real, y::Real)

Return log(exp(x) + exp(y)), avoiding intermediate overflow/undeflow, and handling non-finite values.

StatsFuns.logexpm1 — Function.

logexpm1(x::Real)

Return log(exp(x) - 1) or the "invsoftplus" function. It is the inverse of log1pexp (aka "softplus").

StatsFuns.loghalf

StatsFuns.logistic — Function.

logistic(x::Real)

The logistic sigmoid function mapping a real number to a value in the interval [0,1],

\[\sigma(x) = \frac{1}{e^{-x} + 1} = \frac{e^x}{1+e^x}.\]

Its inverse is the logit function.

StatsFuns.logit — Function.

logit(p::Real)

The logit or log-odds transformation,

\[\log\left(\frac{x}{1-x}\right), \text{where} 0 < x < 1\]

Its inverse is the logistic function.

StatsFuns.logmvgamma — Function.

logmvgamma(p::Int, a::Real)

Return the logarithm of multivariate gamma function (DLMF 35.3.1).

StatsFuns.logmxp1 — Function.

logmxp1(x::Float64)

Return log(x) - x + 1 carefully evaluated.

StatsFuns.logsumexp — Function.

logsumexp(X)

Compute log(sum(exp, X)), evaluated avoiding intermediate overflow/undeflow.

X should be an iterator of real numbers.

StatsFuns.logtwo

StatsFuns.logπ

StatsFuns.lstirlingF32

StatsFuns.lstirlingF64

StatsFuns.lstirling_asym — Function.

lstirling_asym(x)

The remainder term after Stirling's approximation to lgamma:

\[\log \Gamma(x) \approx x \log(x) - x + \log(2\pi/x)/2 = \log(x)*(x-1/2) + \log(2\pi)/2 - x\]

In Julia syntax, this means:

lstirling_asym(x) = lgamma(x) + x - (x-0.5)*log(x) - 0.5*log(2π)

For sufficiently large x, this can be approximated using the asymptotic Stirling's series (DLMF 5.11.1):

\[\frac{1}{12x} - \frac{1}{360x^3} + \frac{1}{1260x^5} - \frac{1}{1680x^7} + \ldots\]

The truncation error is bounded by the first omitted term, and is of the same sign.

Relative error of approximation is bounded by (174611/125400 x^-19) / (1/12 x^-1 - 1/360 x^-3) which is < 1/2 ulp for x >= 10.0, and total numeric error appears to be < 2 ulps

References

Temme, N. (1996) Special functions: An introduction to the classical functions of mathematical physics, Wiley, New York, ISBN: 0-471-11313-1, Chapter 3.6, pp 61-65.

Weisstein, Eric W. "Stirling's Series.". MathWorld.

OEIS A046968 and OEIS A046969 for the series coefficients

StatsFuns.nbetaccdf

StatsFuns.nbetacdf

StatsFuns.nbetainvccdf

StatsFuns.nbetainvcdf

StatsFuns.nbetainvlogccdf

StatsFuns.nbetainvlogcdf

StatsFuns.nbetalogccdf

StatsFuns.nbetalogcdf

StatsFuns.nbetalogpdf

StatsFuns.nbetapdf

StatsFuns.nbinomccdf

StatsFuns.nbinomcdf

StatsFuns.nbinominvccdf

StatsFuns.nbinominvcdf

StatsFuns.nbinominvlogccdf

StatsFuns.nbinominvlogcdf

StatsFuns.nbinomlogccdf

StatsFuns.nbinomlogcdf

StatsFuns.nbinomlogpdf

StatsFuns.nbinompdf

StatsFuns.nchisqccdf

StatsFuns.nchisqcdf

StatsFuns.nchisqinvccdf

StatsFuns.nchisqinvcdf

StatsFuns.nchisqinvlogccdf

StatsFuns.nchisqinvlogcdf

StatsFuns.nchisqlogccdf

StatsFuns.nchisqlogcdf

StatsFuns.nchisqlogpdf

StatsFuns.nchisqpdf

StatsFuns.nfdistccdf

StatsFuns.nfdistcdf

StatsFuns.nfdistinvccdf

StatsFuns.nfdistinvcdf

StatsFuns.nfdistinvlogccdf

StatsFuns.nfdistinvlogcdf

StatsFuns.nfdistlogccdf

StatsFuns.nfdistlogcdf

StatsFuns.nfdistlogpdf

StatsFuns.nfdistpdf

StatsFuns.normccdf

StatsFuns.normcdf

StatsFuns.norminvccdf

StatsFuns.norminvcdf

StatsFuns.norminvlogccdf

StatsFuns.norminvlogcdf

StatsFuns.normlogccdf

StatsFuns.normlogcdf

StatsFuns.normlogpdf

StatsFuns.normpdf

StatsFuns.ntdistccdf

StatsFuns.ntdistcdf

StatsFuns.ntdistinvccdf

StatsFuns.ntdistinvcdf

StatsFuns.ntdistinvlogccdf

StatsFuns.ntdistinvlogcdf

StatsFuns.ntdistlogccdf

StatsFuns.ntdistlogcdf

StatsFuns.ntdistlogpdf

StatsFuns.ntdistpdf

StatsFuns.poisccdf

StatsFuns.poiscdf

StatsFuns.poisinvccdf

StatsFuns.poisinvcdf

StatsFuns.poisinvlogccdf

StatsFuns.poisinvlogcdf

StatsFuns.poislogccdf

StatsFuns.poislogcdf

StatsFuns.poislogpdf

StatsFuns.poispdf

StatsFuns.quartπ

StatsFuns.softmax

StatsFuns.softmax! — Function.

softmax!(r::AbstractArray, x::AbstractArray)

Overwrite r with the softmax (or normalized exponential) transformation of x

That is, r is overwritten with exp.(x), normalized to sum to 1.

See the Wikipedia entry

softmax(x::AbstractArray{<:Real})

Return the softmax transformation applied to x

StatsFuns.softplus — Function.

log1pexp(x::Real)

Return log(1+exp(x)) evaluated carefully for largish x.

This is also called the "softplus" transformation, being a smooth approximation to max(0,x). Its inverse is logexpm1.

StatsFuns.sqrt2

StatsFuns.sqrt2π

StatsFuns.sqrt3

StatsFuns.sqrt4π

StatsFuns.sqrthalfπ

StatsFuns.sqrtπ

StatsFuns.srdistccdf

StatsFuns.srdistcdf

StatsFuns.srdistinvccdf

StatsFuns.srdistinvcdf

StatsFuns.srdistinvlogccdf

StatsFuns.srdistinvlogcdf

StatsFuns.srdistlogccdf

StatsFuns.srdistlogcdf

StatsFuns.tdistccdf

StatsFuns.tdistcdf

StatsFuns.tdistinvccdf

StatsFuns.tdistinvcdf

StatsFuns.tdistinvlogccdf

StatsFuns.tdistinvlogcdf

StatsFuns.tdistlogccdf

StatsFuns.tdistlogcdf

StatsFuns.tdistlogpdf

StatsFuns.tdistpdf

StatsFuns.twoinvπ

StatsFuns.twoπ

StatsFuns.xlogx — Function.

xlogx(x::Real)

Return x * log(x) for x ≥ 0, handling x = 0 by taking the downward limit.

julia> StatsFuns.xlogx(0)
0.0

StatsFuns.xlogy — Function.

xlogy(x::Real, y::Real)

Return x * log(y) for y > 0 with correct limit at x = 0.

StatsFuns.xval

StatsFuns.zval