Docstrings

Atom.@!

Atom.@errs

Atom.@ierrs — Macro.

@ierrs(expression)

Catches errors internal to Atom.jl and renders them in the client.

Atom.@msg — Macro.

@msg(expression)

Calls msg with the name of the top-level function in expression as the first argument and the arguments of that function as the remaining arguments. Note that the toplevel function call will not be evaluated.

Example: @msg(sin(3)) will call msg("sin", 3), i.e.

macroexpand(:(Atom.@msg(sin(3)))) == :((Atom.msg)("sin",3))

Atom.@rpc — Macro.

@rpc(expression)

Calls rpc with the name of the top-level function in expression as the first argument and the arguments of that function as the remaining arguments. Note that the toplevel function call will not be evaluated.

Example: @rpc(sin(3)) will call rpc("sin", 3), i.e.

macroexpand(:(Atom.@rpc(sin(3)))) == :((Atom.rpc)("sin",3))

Atom.@trace

Atom.Atom

Atom.DisplayError

Atom.EvalError

Atom.JunoDisplay

Atom.PlotPaneEnabled

Atom.Profiler

Atom.Progress

Atom.SELECTORS

Atom.__init__

Atom._id

Atom.ans

Atom.appendline

Atom.basecompletionadapter

Atom.baselink

Atom.basepath

Atom.bestlabel

Atom.callbacks

Atom.changeREPLmodule

Atom.changeREPLprompt

Atom.clearconsole

Atom.cliptrace

Atom.completion

Atom.completionmodule

Atom.completionsummary

Atom.completiontext

Atom.completiontype

Atom.connect

Atom.current_prompt

Atom.customdisplaystack

Atom.defaultrepr

Atom.description

Atom.display_error

Atom.display_parse_error

Atom.displayandrender

Atom.displayinplotpane

Atom.ends_with_semicolon

Atom.enter

Atom.eval

Atom.evallock

Atom.evalrepl

Atom.exit_on_sigint

Atom.expandpath

Atom.finddevpackages

Atom.fixdisplayorder

Atom.fixjunodisplays

Atom.fullREPLpath

Atom.fullpath

Atom.generateTreeView

Atom.get_main_mode

Atom.getdocs

Atom.getfield′

Atom.getmethods

Atom.getmoduleinfo

Atom.getmodule′

Atom.gettrace

Atom.gotoitem

Atom.handle — Function.

handle(func, typ)

Set the handler for messages of type typ to func.

Atom.handlemsg — Function.

handlemsg(typ, args...)

Tries to call the message handler corresponding to typ (which can be set with handle).

Atom.handlers

Atom.hideprompt

Atom.highlights

Atom.id

Atom.inREPL

Atom.include

Atom.info

Atom.initialise — Function.

initialise(; welcome = false)

Sets up the environment for Atom.jl: Stop SIGINTs from killing Julia and send a welcome message to Atom if welcome == true.

Atom.inline_mime

Atom.input

Atom.interpose

Atom.isREPL

Atom.isactive — Function.

isactive(sock)

Check whether the socked sock exists and is open.

Atom.isanon

Atom.isconnected

Atom.isdebugging

Atom.isfile′

Atom.ismacro

Atom.ispackage

Atom.isselection

Atom.isuntitled

Atom.jlpane_mime

Atom.juliaprompt

Atom.latex2katex

Atom.locationshading

Atom.matchesprefix

Atom.md_hlines

Atom.methodarray

Atom.moduleinfo

Atom.modulenames

Atom.modulesymbols

Atom.msg — Function.

msg(typ, args...)

Prints a json string to sock that contains all function arguments in an array.

Atom.packageinfo

Atom.pkgpath

Atom.plain_mimes

Atom.plotpane_io_ctx

Atom.plotpane_mime

Atom.plotsize

Atom.ploturl

Atom.realpath′

Atom.renderMD

Atom.renderMDinline

Atom.renderbt

Atom.rendererr

Atom.renderitem

Atom.render′

Atom.repleval

Atom.reset_repl_history

Atom.restoredisplaystack

Atom.returntype

Atom.rpc — Function.

rpc(typ, args...)

Sends a message to sock via msg. Blocks until the client returns a message (which needs to send back the callback field).

Atom.selector

Atom.sendnotify

Atom.serve — Function.

serve(port; kws...)

Start an asynchronous Julia server that listens on localhost:port and handle the messages sent over that socket. The kws... are passed through to initialise.

Atom.showmethod

Atom.showtrace

Atom.sock — Constant.

sock

Socket over which the communication between server and client happens. Initialized by serve.

Atom.stripparams

Atom.strlimit

Atom.structure

Atom.syntaxcolors

Atom.transformwarn

Atom.treeid

Atom.trees

Atom.trim

Atom.typ

Atom.withpath

Atom.wsicon

Atom.wsitem

Atom.wsnamed

Atom.wstype