Docstrings

UnsteadyFlowSolvers.BendingDef

UnsteadyFlowSolvers.ConstDef

UnsteadyFlowSolvers.CosDef

UnsteadyFlowSolvers.EldRampReturnDef

UnsteadyFlowSolvers.EldUpDef

UnsteadyFlowSolvers.EldUpIntDef

UnsteadyFlowSolvers.EldUpInttstartDef

UnsteadyFlowSolvers.EldUptstartDef

UnsteadyFlowSolvers.KelvinCondition

UnsteadyFlowSolvers.KelvinConditionMult

UnsteadyFlowSolvers.KelvinKutta

UnsteadyFlowSolvers.KelvinKuttaMult

UnsteadyFlowSolvers.KinemDef

UnsteadyFlowSolvers.KinemPar

UnsteadyFlowSolvers.KinemPar2DOF

UnsteadyFlowSolvers.LinearDef

UnsteadyFlowSolvers.MotionDef

UnsteadyFlowSolvers.SinDef

UnsteadyFlowSolvers.StepGustDef

UnsteadyFlowSolvers.TwoDFlowField

UnsteadyFlowSolvers.TwoDOFPar

UnsteadyFlowSolvers.TwoDSurf

UnsteadyFlowSolvers.TwoDVort — Type.

TwoDVort(x,z,s,vc,vx,vz)

Defines a 2D vortex at (x,z) with vorticity s and vortex-core radius vc.

vx and vz are induced velocity at centroid used in vortex interaction calculations

UnsteadyFlowSolvers.UnsteadyFlowSolvers

UnsteadyFlowSolvers.add_indbound_b

UnsteadyFlowSolvers.calc_delcp

UnsteadyFlowSolvers.calc_edgeVel

UnsteadyFlowSolvers.calc_forces

UnsteadyFlowSolvers.camber_calc

UnsteadyFlowSolvers.camber_thick_calc

UnsteadyFlowSolvers.checkConverge

UnsteadyFlowSolvers.cleanWrite — Function.

cleanWrite()

Clears all timestamp directories in the current folder

UnsteadyFlowSolvers.controlVortCount — Function.

controlVortCount(delvort, surf, curfield)

Performs merging or deletion operation on free vortices in order to control computational cost according to algorithm specified.

Algorithms for parameter delvort

delNone

Does nothing, no vortex count control.

delSpalart(limit=500, dist=10, tol=1e-6)

Merges vortices according to algorithm given in Spalart, P. R. (1988). Vortex methods for separated flows.

limit: min number of vortices present for merging to occur

dist: small values encourage mergin near airfoil, large values in wake (see paper)

tol: tolerance for merging, merging is less likely to occur for low values (see paper)

There is no universal set of parameters that work for all problem. If using vortex control, test and calibrate choice of parameters

UnsteadyFlowSolvers.delNone

UnsteadyFlowSolvers.delSpalart

UnsteadyFlowSolvers.delVortDef

UnsteadyFlowSolvers.eval

UnsteadyFlowSolvers.find_tstep — Function.

find_tstep([kin])

kin is a definition of either Sin / Cos / Ramp kinematics

A suitable time-step is calculated based on the values of amplitude and frequency

For Sin / Cos kinematics, time step for amp*k = 0.2 is 0.015

For Eldredge ramp kinematics, time step for K = 0.2 is 0.015

For Constant kinematics, time step is 0.015

UnsteadyFlowSolvers.getEndCycle

UnsteadyFlowSolvers.include

UnsteadyFlowSolvers.ind_vel

UnsteadyFlowSolvers.lautat — Function.

lautat(surf, curfield, nsteps=500, dtstar=0.015, startflag=0,

writeflag=0, writeInterval=1000., delvort=delNone(), maxwrite=100, nround=6)

Simulates potential flow for an airfoil undergoing unsteady motion using Large-Angle Unsteady Thin-Airfoil Theory.

Inputs required:
- `TwoDSurf` : based on airfoil geometry and kinematics
- `TwoDFlowField` : initial flowfield

`dtstar` should be in accordance with the kinematics; see `?find_tstep`.

`nsteps` should be calculated according to total run time required and
dtstar.

`startflag=1` can be used to continue a simulation using the
*resultsSummary* file in current directory.

`writeflag=1` along with a suitable writeInterval is used to write flow
data at intervals. These can be used to create plots and movies of the
flowfield.

`delvort` is used to control the vortex count. The simulation cost
increases as O(n``^2``) with number of vortices in the flow field and so
long run times can get very computationally expensive. Look at
`?delVortDef` for algorthims to control vortex count.

`maxwrite` is the maximum number of timestamps/directories that may be
written.

`nround` is the number of digits to which the output timestamps are
rounded off to.

The Large-Angle Unsteady Thin-Airfoil theory is described in: Ramesh,
K. et al., "An unsteady airfoil theory applied to pitching motions
validated against experiment and computation", Theor. Comput. Fluid
Dyn. (2013) 27: 843. [Weblink](https://doi.org/10.1007/s00162-012-0292-8)

UnsteadyFlowSolvers.ldvm

UnsteadyFlowSolvers.ldvm2DOF

UnsteadyFlowSolvers.ldvmLin

UnsteadyFlowSolvers.makeForcePlots2D

UnsteadyFlowSolvers.makeKinemClVortPlots2D

UnsteadyFlowSolvers.makeKinemVelVortPlots2D

UnsteadyFlowSolvers.makeVortPlots2D

UnsteadyFlowSolvers.mutual_ind

UnsteadyFlowSolvers.place_lev

UnsteadyFlowSolvers.place_tev

UnsteadyFlowSolvers.simpleInterp — Function.

simpleInterp(x1, x2, y1, y2, x)

Performs a linear interpolation between points (x1,y1) and (x2,y2) to find the value of y at x.

UnsteadyFlowSolvers.simpleTrapz

UnsteadyFlowSolvers.update_a0anda1

UnsteadyFlowSolvers.update_a2a3adot

UnsteadyFlowSolvers.update_a2toan

UnsteadyFlowSolvers.update_adot

UnsteadyFlowSolvers.update_atermdot

UnsteadyFlowSolvers.update_boundpos

UnsteadyFlowSolvers.update_bv

UnsteadyFlowSolvers.update_downwash

UnsteadyFlowSolvers.update_externalvel

UnsteadyFlowSolvers.update_indbound

UnsteadyFlowSolvers.update_kinem

UnsteadyFlowSolvers.update_kinem2DOF

UnsteadyFlowSolvers.viewVort2D

UnsteadyFlowSolvers.viewVortConnect2D

UnsteadyFlowSolvers.wakeroll

UnsteadyFlowSolvers.writeStamp