Docstrings

Roots.A42 — Type.

Roots.A42()

Bracketing method which finds the root of a continuous function within a provided interval [a, b], without requiring derivatives. It is based on algorithm 4.2 described in: 1. G. E. Alefeld, F. A. Potra, and Y. Shi, "Algorithm 748: enclosing zeros of continuous functions," ACM Trans. Math. Softw. 21, 327–344 (1995), DOI: 10.1145/210089.210111. Originally by John Travers

Roots.AbstractAlefeldPotraShi

Roots.AbstractBisection

Roots.AbstractBracketing

Roots.AbstractHalleyLikeMethod

Roots.AbstractNonBracketing

Roots.AbstractSecant

Roots.AbstractTracks

Roots.AbstractUnivariateZeroMethod

Roots.AbstractUnivariateZeroState

Roots.AlefeldPotraShi — Type.

Roots.AlefeldPotraShi()

Follows algorithm in "ON ENCLOSING SIMPLE ROOTS OF NONLINEAR EQUATIONS", by Alefeld, Potra, Shi; DOI: 10.1090/S0025-5718-1993-1192965-2 link. Efficiency is 1.618. Less efficient, but can be faster than A42() method.

Roots.Bisection — Type.

Bisection()

If possible, will use the bisection method over Float64 values. The bisection method starts with a bracketing interval [a,b] and splits it into two intervals [a,c] and [c,b], If c is not a zero, then one of these two will be a bracketing interval and the process continues. The computation of c is done by _middle, which reinterprets floating point values as unsigned integers and splits there. This method avoids floating point issues and when the tolerances are set to zero (the default) guarantees a "best" solution (one where a zero is found or the bracketing interval is of the type [a, nextfloat(a)]).

When tolerances are given, this algorithm terminates when the midpoint is approximately equal to an endpoint using absolute tolerance xatol and relative tolerance xrtol.

When a zero tolerance is given and the values are not Float64 values, this will call the A42 method.

Roots.BisectionExact

Roots.Brent — Type.

Roots.Brent()

An implementation of Brent's (or Brent-Dekker) method. This method uses a choice of inverse quadratic interpolation or a secant step, falling back on bisection if necessary.

Roots.CallableFunction

Roots.ConvergenceFailed

Roots.DerivativeFree

Roots.Esser — Type.

Order2B()

Esser's method. This is a quadratically convergent method that, like Schroder's method, does not depend on the multiplicity of the zero. Schroder's method has update step x- r2/(r2-r1) * r1, where ri = f^(i-1)/f^(i). Esser approximates f' ~ f[x-h, x+h], f'' ~ f[x-h,x,x+h], where h = fx, as with Steffensen's method, Requiring 3 function calls per step. This implementation uses a secant step when |fx| is considered too large.

Esser, H. Computing (1975) 14: 367. https://doi.org/10.1007/BF02253547 Eine stets quadratisch konvergente Modifikation des Steffensen-Verfahrens

Example

f(x) = cos(x) - x
g(x) = f(x)^2
x0 = pi/4
find_zero(f, x0, Order2(), verbose=true)        #  3 steps / 7 function calls
find_zero(f, x0, Roots.Order2B(), verbose=true) #  4 / 9
find_zero(g, x0, Order2(), verbose=true)        #  22 / 45
find_zero(g, x0, Roots.Order2B(), verbose=true) #  4 / 10

Roots.FalsePosition — Type.

FalsePosition()

Use the false position method to find a zero for the function f within the bracketing interval [a,b].

The false position method is a modified bisection method, where the midpoint between [a_k, b_k] is chosen to be the intersection point of the secant line with the x axis, and not the average between the two values.

To speed up convergence for concave functions, this algorithm implements the 12 reduction factors of Galdino (A family of regula falsi root-finding methods). These are specified by number, as in FalsePosition(2) or by one of three names FalsePosition(:pegasus), FalsePosition(:illinois), or FalsePosition(:anderson_bjork) (the default). The default choice has generally better performance than the others, though there are exceptions.

For some problems, the number of function calls can be greater than for the bisection64 method, but generally this algorithm will make fewer function calls.

Examples

find_zero(x -> x^5 - x - 1, [-2, 2], FalsePosition())

Roots.FirstDerivative

Roots.FloatNN

Roots.FnWrapper

Roots.Halley — Type.

Roots.Halley()

Implements Halley's method, `x_n1 = xn - f/f' * (1 - f/f' * f''/f' * 1/2)^(-1) This method is cubically converging, but requires more function calls per step (3) than other methods.

Example

find_zero((sin, cos, x->-sin(x)), 3.0, Roots.Halley())

If function evaluations are expensive one can pass in a function which returns (f, f/f',f'/f'') as follows

find_zero(x -> (sin(x), sin(x)/cos(x), -cos(x)/sin(x)), 3.0, Roots.Halley())

This can be advantageous if the derivatives are easily computed from the value of f, but otherwise would be expensive to compute.

Roots.Interval

Roots.King — Type.

King()
Order1B()

A superlinear (order 1.6...) modification of the secant method for multiple roots. Presented in A SECANT METHOD FOR MULTIPLE ROOTS, by RICHARD F. KING, BIT 17 (1977), 321-328

The basic idea is similar to Schroder's method: apply the secant method to f/f'. However, this uses f' ~ fp = (fx - f(x-fx))/fx (a Steffensen step). In this implementation, when fx is too big, a single secant step of f is used.

Roots.Newton — Type.

Roots.Newton()

Implements Newton's method: x_n1 = xn - f(xn)/f'(xn). This is a quadratically converging method requiring one derivative. Two function calls per step.

Example

find_zero((sin,cos), 3.0, Roots.Newton())

If function evaluations are expensive one can pass in a function which returns (f, f/f') as follows

find_zero(x -> (sin(x), sin(x)/cos(x)), 3.0, Roots.Newton())

This can be advantageous if the derivative is easily computed from the value of f, but otherwise would be expensive to compute.

Roots.NullTracks

Roots.Order0 — Type.

Order0()

The Order0 method is engineered to be a more robust, though possibly slower, alternative to to the other derivative-free root-finding methods. The implementation roughly follows the algorithm described in Personal Calculator Has Key to Solve Any Equation f(x) = 0, the SOLVE button from the HP-34C. The basic idea is to use a secant step. If along the way a bracket is found, switch to bisection, using AlefeldPotraShi. If the secant step fails to decrease the function value, a quadratic step is used up to 3 times.

Roots.Order1 — Type.

Order1()

The Order1() method is an alias for Secant. It specifies the secant method. This method keeps two values in its state, x_n and x_n1. The updated point is the intersection point of x axis with the secant line formed from the two points. The secant method uses 1 function evaluation per step and has order (1+sqrt(5))/2.

Roots.Order16 — Type.

Order16()

Implement the algorithm from New Sixteenth-Order Derivative-Free Methods for Solving Nonlinear Equations by R. Thukral, American Journal of Computational and Applied Mathematics p-ISSN: 2165-8935; e-ISSN: 2165-8943; 2012; 2(3): 112-118 doi: 10.5923/j.ajcam.20120203.08.

Five function calls per step are required. Though rapidly converging, this method generally isn't faster (fewer function calls/steps) over other methods when using Float64 values, but may be useful for solving over BigFloat.

Roots.Order1B — Type.

King()
Order1B()

A superlinear (order 1.6...) modification of the secant method for multiple roots. Presented in A SECANT METHOD FOR MULTIPLE ROOTS, by RICHARD F. KING, BIT 17 (1977), 321-328

The basic idea is similar to Schroder's method: apply the secant method to f/f'. However, this uses f' ~ fp = (fx - f(x-fx))/fx (a Steffensen step). In this implementation, when fx is too big, a single secant step of f is used.

Roots.Order2 — Type.

Order2()

The quadratically converging Steffensen method is used for the derivative-free Order2() algorithm. Unlike the quadratically converging Newton's method, no derivative is necessary, though like Newton's method, two function calls per step are. Steffensen's algorithm is more sensitive than Newton's method to poor initial guesses when f(x) is large, due to how f'(x) is approximated. This algorithm replaces a Steffensen step with a secant step when f(x) is large.

Roots.Order2B — Type.

Order2B()

Esser's method. This is a quadratically convergent method that, like Schroder's method, does not depend on the multiplicity of the zero. Schroder's method has update step x- r2/(r2-r1) * r1, where ri = f^(i-1)/f^(i). Esser approximates f' ~ f[x-h, x+h], f'' ~ f[x-h,x,x+h], where h = fx, as with Steffensen's method, Requiring 3 function calls per step. This implementation uses a secant step when |fx| is considered too large.

Esser, H. Computing (1975) 14: 367. https://doi.org/10.1007/BF02253547 Eine stets quadratisch konvergente Modifikation des Steffensen-Verfahrens

Example

f(x) = cos(x) - x
g(x) = f(x)^2
x0 = pi/4
find_zero(f, x0, Order2(), verbose=true)        #  3 steps / 7 function calls
find_zero(f, x0, Roots.Order2B(), verbose=true) #  4 / 9
find_zero(g, x0, Order2(), verbose=true)        #  22 / 45
find_zero(g, x0, Roots.Order2B(), verbose=true) #  4 / 10

Roots.Order5 — Type.

Order5()

Implements an order 5 algorithm from A New Fifth Order Derivative Free Newton-Type Method for Solving Nonlinear Equations by Manoj Kumar, Akhilesh Kumar Singh, and Akanksha, Appl. Math. Inf. Sci. 9, No. 3, 1507-1513 (2015), DOI: 10.12785/amis/090346. Four function calls per step are needed.

Roots.Order8 — Type.

Order8()

Implements an eighth-order algorithm from New Eighth-Order Derivative-Free Methods for Solving Nonlinear Equations by Rajinder Thukral, International Journal of Mathematics and Mathematical Sciences Volume 2012 (2012), Article ID 493456, 12 pages DOI: 10.1155/2012/493456. Four function calls per step are required.

Roots.Roots

Roots.Schroder — Type.

Roots.Schroder()

Schroder's method, like Halley's method, utilizes f, f', and f''. Unlike Halley it is quadratically converging, but this is independent of the multiplicity of the zero (cf. Schröder, E. "Über unendlich viele Algorithmen zur Auflösung der Gleichungen." Math. Ann. 2, 317-365, 1870; http://mathworld.wolfram.com/SchroedersMethod.html).

Example

m = 2
f(x) = (cos(x)-x)^m
fp(x) = (-x + cos(x))*(-2*sin(x) - 2)
fpp(x) = 2*((x - cos(x))*cos(x) + (sin(x) + 1)^2)
find_zero((f, fp, fpp), pi/4, Roots.Halley())    # 14 steps
find_zero((f, fp, fpp), 1.0, Roots.Schroder())    # 3 steps

(Whereas, when m=1, Halley is 2 steps to Schroder's 3.)

If function evaluations are expensive one can pass in a function which returns (f, f/f',f'/f'') as follows

find_zero(x -> (sin(x), sin(x)/cos(x), -cos(x)/sin(x)), 3.0, Roots.Schroder())

This can be advantageous if the derivatives are easily computed from the value of f, but otherwise would be expensive to compute.

Roots.Secant — Type.

Order1()

The Order1() method is an alias for Secant. It specifies the secant method. This method keeps two values in its state, x_n and x_n1. The updated point is the intersection point of x axis with the secant line formed from the two points. The secant method uses 1 function evaluation per step and has order (1+sqrt(5))/2.

Roots.SecondDerivative

Roots.StateConverged

Roots.Steffensen

Roots.Tracks

Roots.TupleWrapper — Type.

newton((f, f'), x0; xatol=nothing, xrtol=nothing, maxevals=100)
newton(fΔf, x0; xatol=nothing, xrtol=nothing, maxevals=100)

Newton's method.

Function may be passed in as a tuple (f, f') or as function which returns (f,f/f').

Examples:

newton((sin, cos), 3.0)
newton(x -> (sin(x), sin(x)/cos(x)), 3.0, xatol=1e-10, xrtol=1e-10)

Note: unlike the call newton(f, fp, x0)–which dispatches to a method of find_zero, these two interfaces will specialize on the function that is passed in. This means, these functions will be faster for subsequent calls, but may be slower for an initial call.

Convergence here is decided by xn ≈ x{n-1} using the tolerances specified, which both default to eps(T)^4/5 in the appropriate units.

Roots.UnivariateZeroOptions

Roots.UnivariateZeroState

Roots.__middle

Roots._callable_function

Roots._default_secant_step

Roots._fbracket

Roots._fbracket_diff

Roots._fbracket_ratio

Roots._fz

Roots._fz!

Roots._is_f_approx_0

Roots._method_lookup

Roots._middle

Roots._non_zero

Roots._unitless

Roots.adjust_bracket

Roots.approx_close

Roots.assess_convergence — Function.

Roots.assess_convergence(method, state, options)

Assess if algorithm has converged.

If alogrithm hasn't converged returns false.

If algorithm has stopped or converged, return true and sets one of state.stopped, state.x_converged, state.f_converged, or state.convergence_failed; as well, a message may be set.

state.x_converged = true if abs(xn1 - xn0) < max(xatol, max(abs(xn1), abs(xn0)) * xrtol)

state.f_converged = true if |f(xn1)| < max(atol, |xn1|*rtol)

state.convergence_failed = true if xn1 or fxn1 is NaN or an infinity

state.stopped = true if the number of steps exceed maxevals or the number of function calls exceeds maxfnevals.

In find_zero, stopped values (and x_converged) are checked for convergence with a relaxed tolerance.

Roots.bisection — Function.

bisection(f, a, b; [xatol, xrtol])

Performs bisection method to find a zero of a continuous function.

It is assumed that (a,b) is a bracket, that is, the function has different signs at a and b. The interval (a,b) is converted to floating point and shrunk when a or b is infinite. The function f may be infinite for the typical case. If f is not continuous, the algorithm may find jumping points over the x axis, not just zeros.

If non-trivial tolerances are specified, the process will terminate when the bracket (a,b) satisfies isapprox(a, b, atol=xatol, rtol=xrtol). For zero tolerances, the default, for Float64, Float32, or Float16 values, the process will terminate at a value x with f(x)=0 or f(x)*f(prevfloat(x)) < 0 or f(x) * f(nextfloat(x)) < 0. For other number types, the A42 method is used.

Roots.bisection64

Roots.bracket

Roots.bracketing_error

Roots.callable_function

Roots.check_zero

Roots.choose_smallest

Roots.decide_convergence

Roots.default_tolerances — Function.

default_tolerances(Method, [T], [S])

The default tolerances for most methods are xatol=zero(T), xrtol=eps(T), atol=4eps(S), and rtol=4eps(S), with the proper units (absolute tolerances have the units of x and f(x); relative tolerances are unitless). For Complex{T} values, T is used.

The number of iterations is limited by maxevals=40, the number of function evaluations is not capped, due to maxfnevals=typemax(Int), and strict=false.

default_tolerances(M, [T], [S])

For Bisection (or BisectionExact), when the x values are of type Float64, Float32, or Float16, the default tolerances are zero and there is no limit on the number of iterations or function evalutions. In this case, the algorithm is guaranteed to converge to an exact zero, or a point where the function changes sign at one of the answer's adjacent floating point values.

For other types, the the A42 method (with its tolerances) is used.

default_tolerances(::AbstractAlefeldPotraShi, T, S)

The default tolerances for Alefeld, Potra, and Shi methods are xatol=zero(T), xrtol=2eps(T), atol= zero(S), and rtol=zero(S), with appropriate units; maxevals=45, maxfnevals = Inf; and strict=true.

Roots.derivative_free

Roots.dfree — Function.

dfree(f, xs)

A more robust secant method implementation

Solve for f(x) = 0 using an alogorithm from Personal Calculator Has Key to Solve Any Equation f(x) = 0, the SOLVE button from the HP-34C.

This is also implemented as the Order0 method for find_zero.

The inital values can be specified as a pair of two values, as in (a,b) or [a,b], or as a single value, in which case a value of b is computed, possibly from fb. The basic idea is to follow the secant method to convergence unless:

a bracket is found, in which case bisection is used;

the secant method is not converging, in which case a few steps of a quadratic method are used to see if that improves matters.

Convergence occurs when f(m) == 0, there is a sign change between m and an adjacent floating point value, or f(m) <= 2^3*eps(m).

A value of NaN is returned if the algorithm takes too many steps before identifying a zero.

Examples

Roots.dfree(x -> x^5 - x - 1, 1.0)

Roots.do_steff_step

Roots.eval

Roots.f_ab

Roots.f_abd

Roots.find_non_zero

Roots.find_zero — Function.

find_zero(fs, x0, method; kwargs...)

Interface to one of several methods for find zeros of a univariate function.

Initial starting value

For most methods, x0 is a scalar value indicating the initial value in the iterative procedure. (Secant methods can have a tuple specify their initial values.) Values must be a subtype of Number and have methods for float, real, and oneunit defined.

For bracketing intervals, x0 is specified as a tuple or a vector. A bracketing interval, (a,b), is one where f(a) and f(b) have different signs.

Specifying a method

A method is specified to indicate which algorithm to employ:

There are methods for bisection where a bracket is specified: Bisection, Roots.A42, FalsePosition

There are several derivative-free methods: cf. Order0, Order1 (secant method), Order2 (Steffensen), Order5, Order8, and Order16, where the number indicates the order of the convergence.

There are some classical methods where derivatives are required: Roots.Newton, Roots.Halley. (The are not exported.)

For more detail, see the help page for each method (e.g., ?Order1).

If no method is specified, the default method depends on x0:

If x0 is a scalar, the default is the slower, but more robust Order0 method.

If x0 is a tuple or vector indicating a bracketing interval, the Bisection method is used. (The exact algorithm depends on the number type, the tolerances, and verbose.)

Specifying the function

The function(s) are passed as the first argument.

For the few methods that use a derivative (Newton, Halley, and optionally Order5) a tuple of functions is used.

Optional arguments (tolerances, limit evaluations, tracing)

xatol - absolute tolerance for x values. Passed to isapprox(x_n, x_{n-1})

xrtol - relative tolerance for x values. Passed to isapprox(x_n, x_{n-1})

atol - absolute tolerance for f(x) values.

rtol - relative tolerance for f(x) values.

maxevals - limit on maximum number of iterations

maxfnevals - limit on maximum number of function evaluations

strict - if false (the default), when the algorithm stops, possible zeros are checked with a relaxed tolerance

verbose - if true a trace of the algorithm will be shown on successful completion.

See the help string for Roots.assess_convergence for details on convergence. See the help page for Roots.default_tolerances(method) for details on the default tolerances.

In general, with floating point numbers, convergence must be understood as not an absolute statement. Even if mathematically x is an answer the floating point realization, say xstar, it may be that f(xstar) - f(x) = f(xstar) ≈ f'(x) ⋅ eps(x), so tolerances must be appreciated, and at times specified.

For the Bisection methods, convergence is guaranteed, so the tolerances are set to be 0 by default.

If a bracketing method is passed in after the method specification if a bracket is found, the method will switch. This is what Order0 does by default, with a secant step initially.

Examples:

# default methods
find_zero(sin, 3)  # use Order0()
find_zero(sin, (3,4)) # use Bisection()

# specifying a method
find_zero(sin, 3.0, Order2())              # Use Steffensen method
find_zero(sin, big(3.0), Order16())        # rapid convergence
find_zero(sin, (3, 4), FalsePosition())    # fewer function calls than Bisection(), in this case
find_zero(sin, (3, 4), FalsePosition(8))   # 1 or 12 possible algorithms for false position
find_zero((sin,cos), 3.0, Roots.Newton())  # use Newton's method
find_zero((sin, cos, x->-sin(x)), 3.0, Roots.Halley())  # use Halley's method

# changing tolerances
fn, x0, xstar = (x -> (2x*cos(x) + x^2 - 3)^10/(x^2 + 1), 3.0,  2.9806452794385368)
find_zero(fn, x0, Order2()) - xstar        # 0.014079847201995843
find_zero(fn, x0, Order2(), atol=0.0, rtol=0.0) # error: x_n ≉ x_{n-1}; just f(x_n) ≈ 0
fn, x0, xstar = (x -> (sin(x)*cos(x) - x^3 + 1)^9,        1.0,  1.117078770687451)
find_zero(fn, x0, Order2())                # 1.1122461983100858
find_zero(fn, x0, Order2(), maxevals=3)    # Roots.ConvergenceFailed: 26 iterations needed

# tracing output
find_zero(x->sin(x), 3.0, Order2(), verbose=true)   # 3 iterations
find_zero(x->sin(x)^5, 3.0, Order2(), verbose=true) # 22 iterations

Roots.find_zeros — Function.

findzeros(f, a, b; [nopts=12, k=8, naive=false, xatol, xrtol, atol, rtol])

Search for zeros of f in the interval [a,b].

Examples

find_zeros(x -> exp(x) - x^4, -5, 20)        # a few well-spaced zeros
find_zeros(x -> sin(x^2) + cos(x)^2, 0, 10)  # many zeros
find_zeros(x -> cos(x) + cos(2x), 0, 4pi)    # mix of simple, non-simple zeros
f(x) = (x-0.5) * (x-0.5001) * (x-1)          # nearby zeros
find_zeros(f, 0, 2)
f(x) = (x-0.5) * (x-0.5001) * (x-4) * (x-4.001) * (x-4.2)
find_zeros(f, 0, 10)
f(x) = (x-0.5)^2 * (x-0.5001)^3 * (x-4) * (x-4.001) * (x-4.2)^2  # hard to identify
find_zeros(f, 0, 10, no_pts=21)                # too hard for default

Notes:

There are two typical cases where the number of zeros may be underreported:

if the initial interval, [a,b], is too wide

if there are nearby zeros

The basic algorithm checks for zeros among the endpoints, and then divides the interval (a,b) into no_pts-1 subintervals and then proceeds to look for zeros through bisection or a derivative-free method. As checking for a bracketing interval is relatively cheap and bisection is guaranteed to converge, each interval has k pairs of intermediate points checked for a bracket.

If any zeros are found, the algorithm uses these to partition (a,b) into subintervals. Each subinterval is shrunk so that the endpoints are not zeros and the process is repeated on the subinterval. If the initial interval is too large, then the naive scanning for zeros may be fruitless and no zeros will be reported. If there are nearby zeros, the shrinking of the interval may jump over them, though as seen in the examples, nearby roots can be identified correctly, though for really nearby points, or very flat functions, it may help to increase no_pts.

The tolerances are used to shrink the intervals, but not to find zeros within a search. For searches, bisection is guaranteed to converge with no specified tolerance. For the derivative free search, a modification of the Order0 method is used, which at worst case compares |f(x)| <= 8*eps(x) to identify a zero. The algorithm might identify more than one value for a zero, due to floating point approximations. If a potential pair of zeros satisfy isapprox(a,b,atol=sqrt(xatol), rtol=sqrt(xrtol)) then they are consolidated.

The algorithm can make many function calls. When zeros are found in an interval, the naive search is carried out on each subinterval. To cut down on function calls, though with some increased chance of missing some zeros, the adaptive nature can be skipped with the argument naive=true or the number of points stepped down.

The algorithm is derived from one in a PR by @djsegal.

Roots.fzero — Function.

fzero(f, x0; order=0; kwargs...)
fzero(f, x0, M; kwargs...)
fzero(f, x0, M, N; kwargs...)
fzero(f, x0; kwargs...)
fzero(f, a::Number, b::Numbers; kwargs...)
fzero(f, a::Number, b::Numbers; order=?, kwargs...)
fzero(f, fp, a::Number; kwargs...)

Find zero of a function using one of several iterative algorithms.

f: a scalar function or callable object

x0: an initial guess, a scalar value or tuple of two values

order: An integer, symbol, or string indicating the algorithm to use for find_zero. The Order0 default may be specified directly by order=0, order=:0, or order="0"; Order1() by order=1, order=:1, order="1", or order=:secant; Order1B() by `order="1B", etc.

M: a specific method, as would be passed to find_zero, bypassing the use of the order keyword

N: a specific bracketing method. When given, if a bracket is identified, method N will be used to finish instead of method M.

a, b: When two values are passed along, if no order value is specified, Bisection will be used over the bracketing interval (a,b). If an order value is specified, the value of x0 will be set to (a,b) and the specified method will be used.

fp: when fp is specified (assumed to compute the derivative of f), Newton's method will be used

kwargs...: See find_zero for the specification of tolerances and other keyword arguments

Examples:

fzero(sin, 3)                  # use Order0() method, the default
fzero(sin, 3, order=:secant)   # use secant method (also just `order=1`)
fzero(sin, 3, Roots.Order1B()) # use secant method variant for multiple roots.
fzero(sin, 3, 4)               # use bisection method over (3,4)
fzero(sin, 3, 4, xatol=1e-6)   # use bisection method until |x_n - x_{n-1}| <= 1e-6
fzero(sin, 3, 3.1, order=1)    # use secant method with x_0=3.0, x_1 = 3.1
fzero(sin, (3, 3.1), order=2)  # use Steffensen's method with x_0=3.0, x_1 = 3.1
fzero(sin, cos, 3)             # use Newton's method

Note: unlike find_zero, fzero does not specialize on the type of the function argument. This has the advantage of making the first use of the function f faster, but subsequent uses slower.

Roots.fzeros — Function.

fzeros(f, a, b; kwargs...)

Searches for all zeros of f within an interval (a,b). Assume neither a or b is a zero.

Dispatches to find_zeros(f, a, b; kwargs...).

Roots.fΔx

Roots.fΔxΔΔx

Roots.galdino

Roots.galdino_reduction

Roots.guarded_secant_step

Roots.halley — Function.

Roots.halley(f, fp, fpp, x0; kwargs...)

Implementation of Halley's method (cf ?Roots.Halley()).

Arguments:

f::Function – function to find zero of

fp::Function – derivative of f.

fpp:Function – second derivative of f.

x0::Number – initial guess

With the FowardDiff package derivatives may be computed automatically. For example, defining D(f) = x -> ForwardDiff.derivative(f, float(x)) allows D(f) and D(D(f)) to be used for the first and second derivatives, respectively.

Keyword arguments are passed to find_zero using the Roots.Halley() method.

Roots.has_converged

Roots.identify_starting_point

Roots.incfn

Roots.include

Roots.incsteps

Roots.init_options

Roots.init_options!

Roots.init_state

Roots.init_state!

Roots.ipzero

Roots.isbracket

Roots.isissue

Roots.isneg

Roots.log_step

Roots.make_intervals!

Roots.newton — Function.

Roots.newton(f, fp, x0; kwargs...)

Implementation of Newton's method: x_n1 = x_n - f(x_n)/ f'(x_n)

Arguments:

f::Function – function to find zero of

fp::Function – the derivative of f.

x0::Number – initial guess. For Newton's method this may be complex.

With the FowardDiff package derivatives may be computed automatically. For example, defining D(f) = x -> ForwardDiff.derivative(f, float(x)) allows D(f) to be used for the first derivative.

Keyword arguments are passed to find_zero using the Roots.Newton() method.

See also Roots.newton((f,fp), x0) andRoots.newton(fΔf, x0)` for simpler implementations.

Roots.newton_quadratic

Roots.not_near

Roots.quad_vertex

Roots.run_bisection

Roots.secant

Roots.secant_method — Function.

secant_method(f, xs; [atol=0.0, rtol=8eps(), maxevals=1000])

Perform secant method to solve f(x) = 0.

The secant method is an iterative method with update step given by b - fb/m where m is the slope of the secant line between (a,fa) and (b,fb).

The inital values can be specified as a pair of 2, as in (a,b) or [a,b], or as a single value, in which case a value of b is chosen.

The algorithm returns m when abs(fm) <= max(atol, abs(m) * rtol). If this doesn't occur before maxevals steps or the algorithm encounters an issue, a value of NaN is returned. If too many steps are taken, the current value is checked to see if there is a sign change for neighboring floating point values.

The Order1 method for find_zero also implements the secant method. This one will be faster, as there are fewer setup costs.

Examples:

Roots.secant_method(sin, (3,4))
Roots.secant_method(x -> x^5 -x - 1, 1.1)

Note:

This function will specialize on the function f, so that the inital call can take more time than a call to the Order1() method, though subsequent calls will be much faster. Using FunctionWrappers.jl can ensure that the initial call is also equally as fast as subsequent ones.

Roots.secant_step

Roots.show_trace

Roots.show_tracks

Roots.sort_smallest

Roots.steff_step — Function.

heuristic to get a decent first step with Steffensen steps

Roots.take_a42_step

Roots.update_state