Docstrings

Distances.ArraySlice

Distances.BhattacharyyaDist

Distances.BrayCurtis

Distances.Bregman — Type.

Implements the Bregman divergence, a friendly introduction to which can be found here. Bregman divergences are a minimal implementation of the "mean-minimizer" property.

It is assumed that the (convex differentiable) function F maps vectors (of any type or size) to real numbers. The inner product used is Base.dot, but one can be passed in either by defining inner or by passing in a keyword argument. If an analytic gradient isn't available, Julia offers a suite of good automatic differentiation packages.

function evaluate(dist::Bregman, p::AbstractVector, q::AbstractVector)

Distances.Chebyshev

Distances.ChiSqDist

Distances.Cityblock

Distances.CorrDist

Distances.CosineDist

Distances.Distances

Distances.Euclidean — Type.

Euclidean([thresh])

Create a euclidean metric.

When computing distances among large numbers of points, it can be much more efficient to exploit the formula

(x-y)^2 = x^2 - 2xy + y^2

However, this can introduce roundoff error. thresh (which defaults to 0) specifies the relative square-distance tolerance on 2xy compared to x^2 + y^2 to force recalculation of the distance using the more precise direct (elementwise-subtraction) formula.

Example:

julia> x = reshape([0.1, 0.3, -0.1], 3, 1);

julia> pairwise(Euclidean(), x, x)
1×1 Array{Float64,2}:
 7.45058e-9

julia> pairwise(Euclidean(1e-12), x, x)
1×1 Array{Float64,2}:
 0.0

Distances.GenKLDivergence

Distances.Hamming

Distances.Haversine — Type.

Haversine(radius)

The haversine distance between two locations on a sphere of given radius.

Locations are described with longitude and latitude in degrees. The computed distance has the same units as that of the radius.

Distances.HellingerDist

Distances.JSDivergence

Distances.Jaccard

Distances.KLDivergence

Distances.Mahalanobis

Distances.MeanAbsDeviation

Distances.MeanSqDeviation

Distances.Metric

Distances.Minkowski

Distances.NormRMSDeviation

Distances.PreMetric

Distances.RMSDeviation

Distances.RealAbstractArray

Distances.RenyiDivergence — Type.

RenyiDivergence(α::Real)
renyi_divergence(P, Q, α::Real)

Create a Rényi premetric of order α.

Rényi defined a spectrum of divergence measures generalising the Kullback–Leibler divergence (see KLDivergence). The divergence is not a semimetric as it is not symmetric. It is parameterised by a parameter α, and is equal to Kullback–Leibler divergence at α = 1:

At α = 0, $R_0(P | Q) = -log(sum_{i: p_i > 0}(q_i))$

At α = 1, $R_1(P | Q) = sum_{i: p_i > 0}(p_i log(p_i / q_i))$

At α = ∞, $R_∞(P | Q) = log(sup_{i: p_i > 0}(p_i / q_i))$

Otherwise $R_α(P | Q) = log(sum_{i: p_i > 0}((p_i ^ α) / (q_i ^ (α - 1))) / (α - 1)$

Example:

julia> x = reshape([0.1, 0.3, 0.4, 0.2], 2, 2);

julia> pairwise(RenyiDivergence(0), x, x)
2×2 Array{Float64,2}:
 0.0  0.0
 0.0  0.0

julia> pairwise(Euclidean(2), x, x)
2×2 Array{Float64,2}:
 0.0       0.577315
 0.655407  0.0

Distances.RogersTanimoto

Distances.SemiMetric

Distances.SpanNormDist

Distances.SqEuclidean — Type.

SqEuclidean([thresh])

Create a squared-euclidean semi-metric. For the meaning of thresh, see Euclidean.

Distances.SqMahalanobis

Distances.UnionMetrics

Distances.UnionWeightedMetrics

Distances.VecOrLengthTwoTuple

Distances.WeightedCityblock

Distances.WeightedEuclidean

Distances.WeightedHamming

Distances.WeightedMinkowski

Distances.WeightedSqEuclidean

Distances._centralize

Distances._centralize_colwise

Distances.bhattacharyya

Distances.bhattacharyya_coeff

Distances.braycurtis

Distances.bregman

Distances.chebyshev

Distances.chisq_dist

Distances.cityblock

Distances.colwise

Distances.colwise!

Distances.corr_dist

Distances.cosine_dist

Distances.dot_percol

Distances.dot_percol!

Distances.euclidean

Distances.eval

Distances.eval_end

Distances.eval_op

Distances.eval_reduce

Distances.eval_start

Distances.evaluate

Distances.get_colwise_dims

Distances.get_common_ncols

Distances.get_pairwise_dims

Distances.gkl_divergence

Distances.hamming

Distances.haversine

Distances.haversine_error

Distances.hellinger

Distances.include

Distances.jaccard

Distances.js_divergence

Distances.kl_divergence

Distances.mahalanobis

Distances.meanad

Distances.minkowski

Distances.msd

Distances.nrmsd

Distances.pairwise

Distances.pairwise!

Distances.renyi_divergence — Function.

RenyiDivergence(α::Real)
renyi_divergence(P, Q, α::Real)

Create a Rényi premetric of order α.

Rényi defined a spectrum of divergence measures generalising the Kullback–Leibler divergence (see KLDivergence). The divergence is not a semimetric as it is not symmetric. It is parameterised by a parameter α, and is equal to Kullback–Leibler divergence at α = 1:

At α = 0, $R_0(P | Q) = -log(sum_{i: p_i > 0}(q_i))$

At α = 1, $R_1(P | Q) = sum_{i: p_i > 0}(p_i log(p_i / q_i))$

At α = ∞, $R_∞(P | Q) = log(sup_{i: p_i > 0}(p_i / q_i))$

Otherwise $R_α(P | Q) = log(sum_{i: p_i > 0}((p_i ^ α) / (q_i ^ (α - 1))) / (α - 1)$

Example:

julia> x = reshape([0.1, 0.3, 0.4, 0.2], 2, 2);

julia> pairwise(RenyiDivergence(0), x, x)
2×2 Array{Float64,2}:
 0.0  0.0
 0.0  0.0

julia> pairwise(Euclidean(2), x, x)
2×2 Array{Float64,2}:
 0.0       0.577315
 0.655407  0.0

Distances.result_type

Distances.rmsd

Distances.rogerstanimoto

Distances.spannorm_dist

Distances.sqeuclidean

Distances.sqmahalanobis

Distances.sqrt!

Distances.sumsq_percol

Distances.wcityblock

Distances.weuclidean

Distances.whamming

Distances.wminkowski

Distances.wsqeuclidean

Distances.wsumsq_percol