Docstrings

POMDPs.@POMDP_require — Macro.

@POMDP_require solve(s::CoolSolver, p::POMDP) begin
    PType = typeof(p)
    @req states(::PType)
    @req actions(::PType)
    @req transition(::PType, ::S, ::A)
    s = first(states(p))
    a = first(actions(p))
    t_dist = transition(p, s, a)
    @req rand(::AbstractRNG, ::typeof(t_dist))
end

Create a get_requirements implementation for the function signature and the requirements block.

POMDPs.@POMDP_requirements — Macro.

reqs = @POMDP_requirements CoolSolver begin
    PType = typeof(p)
    @req states(::PType)
    @req actions(::PType)
    @req transition(::PType, ::S, ::A)
    s = first(states(p))
    a = first(actions(p))
    t_dist = transition(p, s, a)
    @req rand(::AbstractRNG, ::typeof(t_dist))
end

Create a RequirementSet object.

POMDPs.@get_requirements — Macro.

@get_requirements f(arg1, arg2)

Call get_requirements(f, (arg1,arg2)).

POMDPs.@impl_dep — Macro.

@impl_dep reward(::P,::S,::A,::S) where {P<:POMDP,S,A} reward(::P,::S,::A)

Declare an implementation dependency and automatically implement implemented.

In the example above, @implemented reward(::P,::S,::A,::S) will return true if the user has implemented reward(::P,::S,::A,::S) OR reward(::P,::S,::A)

THIS IS ONLY INTENDED FOR USE INSIDE POMDPs AND MAY NOT FUNCTION CORRECTLY ELSEWHERE

POMDPs.@implemented — Macro.

@implemented function(::Arg1Type, ::Arg2Type)

Check whether there is an implementation available that will return a suitable value.

POMDPs.@req — Macro.

@req f( ::T1, ::T2)

Convert a f( ::T1, ::T2) expression to a (f, Tuple{T1,T2})::Req for pushing to a RequirementSet.

If in a @POMDP_requirements or @POMDP_require block, marks the requirement for including in the set of requirements.

POMDPs.@requirements_info — Macro.

@requirements_info ASolver() [YourPOMDP()]

Print information about the requirements for a solver.

POMDPs.@show_requirements — Macro.

@show_requirements solve(solver, problem)

Print a a list of requirements for a function call.

POMDPs.@subreq — Macro.

@subreq f(arg1, arg2)

In a @POMDP_requirements or @POMDP_require block, include the requirements for f(arg1, arg2) as a child argument set.

POMDPs.@warn_requirements — Macro.

@warn_requirements solve(solver, problem)

Print a warning if there are unmet requirements.

POMDPs.AbstractRequirementSet

POMDPs.CheckedList

POMDPs.EXPORTED_TYPES

POMDPs.MDP — Type.

Abstract base type for a fully observable Markov decision process.

S: state type
A: action type

POMDPs.NATIVE_PACKAGES — Constant.

Set containing string names of officially supported solvers and utility packages (e.g. MCTS, SARSOP, POMDPToolbox, etc). If you have a validated solver that supports the POMDPs.jl API, contact the developers to add your solver to this list.

POMDPs.NON_NATIVE_PACKAGES

POMDPs.POMDP — Type.

Abstract base type for a partially observable Markov decision process.

S: state type
A: action type
O: observation type

POMDPs.POMDP_REGISTRY — Constant.

url to JuliaPOMDP registry

POMDPs.POMDPs — Module.

Provides a basic interface for defining and solving MDPs/POMDPs

POMDPs.Policy — Type.

Base type for a policy (a map from every possible belief, or more abstract policy state, to an optimal or suboptimal action)

POMDPs.REGISTERED_PACKAGES

POMDPs.REMOTE_URL — Constant.

url to remote JuliaPOMDP organization repo

POMDPs.REQUIRED_REGISTRY_ENTRIES

POMDPs.Req

POMDPs.RequirementSet

POMDPs.SUPPORTED_PACKAGES

POMDPs.Simulator — Type.

Base type for an object defining how simulations should be carried out.

POMDPs.Solver — Type.

Base type for an MDP/POMDP solver

POMDPs.TupleType

POMDPs.Unspecified

POMDPs.Updater — Type.

Abstract type for an object that defines how the belief should be updated

A belief is a general construct that represents the knowledge an agent has about the state of the system. This can be a probability distribution, an action observation history or a more general representation.

POMDPs.action — Function.

action{B}(policy::Policy, x::B)

Fills and returns action based on the current state or belief, given the policy. B is a generalized information state - can be a state in an MDP, a distribution in POMDP, or any other representation needed to make a decision using the given policy.

POMDPs.action_index

POMDPs.action_type

POMDPs.actionindex — Function.

actionindex{S,A,O}(problem::POMDP{S,A,O}, a::A)
actionindex{S,A}(problem::MDP{S,A}, a::A)

Return the integer index of action a. Used for discrete models only.

POMDPs.actions — Function.

actions(problem::POMDP)
actions(problem::MDP)

Returns the entire action space of a POMDP.

POMDPs.actiontype — Function.

actiontype(t::Type)
actiontype(p::Union{POMDP,MDP})

Return the state type for a problem type (the S in POMDP{S,A,O}).

type A <: POMDP{Bool, Int, Bool} end

actiontype(A) # returns Int

POMDPs.add

POMDPs.add_all

POMDPs.add_registry — Function.

add_registry()

Adds the JuliaPOMDP registry

POMDPs.available — Function.

available()

Prints all the available packages in the JuliaPOMDP registry

POMDPs.build

POMDPs.check_requirements — Function.

check_requirements(r::AbstractRequirementSet)

Check whether the methods in r have implementations with implemented(). Return true if all methods have implementations.

POMDPs.convert_a — Function.

convert_a(::Type{V}, a, problem::Union{MDP,POMDP}) where V<:AbstractArray
convert_a(::Type{A}, vec::V, problem::Union{MDP,POMDP}) where {A,V<:AbstractArray}

Convert an action to vectorized form or vice versa.

POMDPs.convert_call — Function.

Return a (f, (arg1,arg2)) expression given a f(arg1, arg2) expression.

POMDPs.convert_o — Function.

convert_o(::Type{V}, o, problem::Union{MDP,POMDP}) where V<:AbstractArray
convert_o(::Type{O}, vec::V, problem::Union{MDP,POMDP}) where {O,V<:AbstractArray}

Convert an observation to vectorized form or vice versa.

POMDPs.convert_req — Function.

Return a (f, Tuple{T1,T2})::Req expression given a f( ::T1, ::T2) expression.

POMDPs.convert_s — Function.

convert_s(::Type{V}, s, problem::Union{MDP,POMDP}) where V<:AbstractArray
convert_s(::Type{S}, vec::V, problem::Union{MDP,POMDP}) where {S,V<:AbstractArray}

Convert a state to vectorized form or vice versa.

POMDPs.dimensions — Function.

dimensions(s::Any)

Returns the number of dimensions in space s.

POMDPs.discount — Function.

discount(problem::POMDP)
discount(problem::MDP)

Return the discount factor for the problem.

POMDPs.eval

POMDPs.failed_synth_warning

POMDPs.format_method

POMDPs.generate_o — Function.

generate_o{S,A,O}(p::POMDP{S,A,O}, s::S, a::A, sp::S, rng::AbstractRNG)

Return the next observation given current state s, action taken a and next state sp.

Usually the observation would only depend on the next state sp.

generate_o{S,A,O}(p::POMDP{S,A,O}, s::S, rng::AbstractRNG)

Return the observation from the current state. This should be used to generate initial observations.

POMDPs.generate_or — Function.

generate_or{S,A,O}(p::POMDP{S,A,O}, s::S, a::A, sp::S, rng::AbstractRNG)

Return the observation o and reward for taking action a in current state s reaching state sp.

POMDPs.generate_s — Function.

generate_s{S,A}(p::Union{POMDP{S,A},MDP{S,A}}, s::S, a::A, rng::AbstractRNG)

Return the next state given current state s and action taken a.

POMDPs.generate_so — Function.

generate_so{S,A,O}(p::POMDP{S,A,O}, s::S, a::A, rng::AbstractRNG)

Return the next state sp and observation o.

POMDPs.generate_sor — Function.

generate_sor{S,A,O}(p::POMDP{S,A,O}, s::S, a::A, rng::AbstractRNG)

Return the next state sp, observation o and reward for taking action a in current state s.

POMDPs.generate_sr — Function.

generate_sr{S}(p::Union{POMDP{S},MDP{S}}, s, a, rng::AbstractRNG)

Return the next state sp and reward for taking action a in current state s.

POMDPs.get_requirements — Function.

get_requirements(f::Function, args::Tuple)

Return a RequirementSet for the function f and arguments args.

POMDPs.getindex

POMDPs.handle_method

POMDPs.handle_reqs! — Function.

handle_reqs!(block, reqs_name::Symbol)

Replace any @req calls with push!($reqs_name, <requirement>)

Returns true if there was a requirement in there and so should not be escaped.

POMDPs.implemented — Function.

implemented(function, Tuple{Arg1Type, Arg2Type})

Check whether there is an implementation available that will return a suitable value.

POMDPs.include

POMDPs.initial_state

POMDPs.initial_state_distribution

POMDPs.initialize_belief — Function.

initialize_belief(updater::Updater,
                     state_distribution::Any)
initialize_belief(updater::Updater, belief::Any)

Returns a belief that can be updated using updater that has similar distribution to state_distribution or belief.

The conversion may be lossy. This function is also idempotent, i.e. there is a default implementation that passes the belief through when it is already the correct type: initialize_belief(updater::Updater, belief) = belief

POMDPs.initialstate — Function.

initialstate{S}(p::Union{POMDP{S},MDP{S}}, rng::AbstractRNG)

Return the initial state for the problem p.

Usually the initial state is sampled from an initial state distribution.

POMDPs.initialstate_distribution — Function.

initialstate_distribution(pomdp::POMDP)
initialstate_distribution(mdp::MDP)

Return a distribution of the initial state of the pomdp or mdp.

POMDPs.isterminal — Function.

isterminal{S,A,O}(problem::POMDP{S,A,O}, state::S)
isterminal{S,A}(problem::MDP{S,A}, state::S)

Check if state s is terminal

POMDPs.isterminal_obs — Function.

isterminal_obs{S,A,O}(problem::POMDP{S,A,O}, observation::O)

Check if an observation is terminal.

POMDPs.iterator — Function.

iterator(d::Any)

DEPRECATED. Return an iterable object (array or custom iterator) that iterates over possible values of distribution d. Values with zero probability may be skipped.

POMDPs.length

POMDPs.lowerbound

Statistics.mean — Function.

mean(d::Any)

Return the mean of a distribution d.

StatsBase.mode — Function.

mode(d::Any)

Return the most likely value in a distribution d.

POMDPs.n_actions — Function.

n_actions(problem::POMDP)
n_actions(problem::MDP)

Return the number of actions in problem. Used for discrete models only.

POMDPs.n_observations — Function.

n_observations(problem::POMDP)

Return the number of observations in problem. Used for discrete models only.

POMDPs.n_states — Function.

n_states(problem::POMDP)
n_states(problem::MDP)

Return the number of states in problem. Used for discrete models only.

POMDPs.obs_index

POMDPs.obs_type

POMDPs.observation — Function.

observation{S,A,O}(problem::POMDP{S,A,O}, statep::S)
observation{S,A,O}(problem::POMDP{S,A,O}, action::A, statep::S)
observation{S,A,O}(problem::POMDP{S,A,O}, state::S, action::A, statep::S)

Return the observation distribution. You need only define the method with the fewest arguments needed to determine the observation distribution.

Example

using POMDPToolbox # for SparseCat

struct MyPOMDP <: POMDP{Int, Int, Int} end

observation(p::MyPOMDP, sp::Int) = SparseCat([sp-1, sp, sp+1], [0.1, 0.8, 0.1])

POMDPs.observations — Function.

observations(problem::POMDP)

Return the entire observation space.

POMDPs.obsindex — Function.

obsindex{S,A,O}(problem::POMDP{S,A,O}, o::O)

Return the integer index of observation o. Used for discrete models only.

POMDPs.obstype — Function.

obstype(t::Type)

Return the state type for a problem type (the S in POMDP{S,A,O}).

type A <: POMDP{Bool, Bool, Int} end

obstype(A) # returns Int

Distributions.pdf — Function.

pdf(d::Any, x::Any)

Evaluate the probability density of distribution d at sample x.

POMDPs.pomdp_requirements — Function.

Return an expression that creates a RequirementSet using the code in the block. The resulting code will always return a RequirementSet, but it may be incomplete if the exception field is not null.

POMDPs.push_dep!

Base.rand — Function.

rand{T}(rng::AbstractRNG, d::Any)

Return a random element from distribution or space d.

If d is a state or transition distribution, the sample will be a state; if d is an action distribution, the sample will be an action or if d is an observation distribution, the sample will be an observation.

POMDPs.read_registry

POMDPs.recursively_check

POMDPs.recursively_show

POMDPs.remove_all

POMDPs.requirements_info — Function.

requirements_info(s::Solver, p::Union{POMDP,MDP}, ...)

Print information about the requirement for solver s.

POMDPs.reward — Function.

reward{S,A,O}(problem::POMDP{S,A,O}, state::S, action::A)
reward{S,A}(problem::MDP{S,A}, state::S, action::A)

Return the immediate reward for the s-a pair

POMDPs.sampletype — Function.

sampletype(T::Type)
sampletype(d::Any) = sampletype(typeof(d))

Return the type of objects that are sampled from a distribution or space d when rand(rng, d) is called.

The distribution writer should implement the sampletype(::Type) method for the distribution type, then the function can be called for that type or for objects of that type (i.e. the sampletype(d::Any) = sampletype(typeof(d)) default is provided).

POMDPs.short_method

POMDPs.show_checked_list

POMDPs.show_heading

POMDPs.show_incomplete

POMDPs.show_requirements — Function.

show_requirements(r::AbstractRequirementSet)

Check whether the methods in r have implementations with implemented() and print out a formatted list showing which are missing. Return true if all methods have implementations.

POMDPs.show_requirer

POMDPs.simulate — Function.

simulate{S,A,O,B}(simulator::Simulator, problem::POMDP{S,A,O}, policy::Policy{B}, updater::Updater{B}, initial_belief::B)
simulate{S,A}(simulator::Simulator, problem::MDP{S,A}, policy::Policy, initial_state::S)

Run a simulation using the specified policy.

The return type is flexible and depends on the simulator. For example implementations, see the POMDPToolbox package.

POMDPs.solve — Function.

solve(solver::Solver, problem::POMDP)

Solves the POMDP using method associated with solver, and returns a policy.

POMDPs.state_index

POMDPs.state_type

POMDPs.stateindex — Function.

stateindex{S,A,O}(problem::POMDP{S,A,O}, s::S)
stateindex{S,A}(problem::MDP{S,A}, s::S)

Return the integer index of state s. Used for discrete models only.

POMDPs.states — Function.

states(problem::POMDP)
states(problem::MDP)

Returns the complete state space of a POMDP.

POMDPs.statetype — Function.

statetype(t::Type)
statetype(p::Union{POMDP,MDP})

Return the state type for a problem type (the S in POMDP{S,A,O}).

type A <: POMDP{Int, Bool, Bool} end

statetype(A) # returns Int

Distributions.support — Function.

support(d::Any)

Return the possible values that can be sampled from distribution d. Values with zero probability may be skipped.

POMDPs.test_all

POMDPs.transition — Function.

transition{S,A,O}(problem::POMDP{S,A,O}, state::S, action::A)
transition{S,A}(problem::MDP{S,A}, state::S, action::A)

Return the transition distribution from the current state-action pair

POMDPs.unpack_typedcall — Function.

Return a tuple (not an Expr) of the function name, arguments, and argument types.

E.g. f(arg1::T1, arg2::T2) would be unpacked to (:f, [:arg1, :arg2], [:T1, :T2])

POMDPs.update — Function.

update(updater::Updater, belief_old, action, observation)

Return a new instance of an updated belief given belief_old and the latest action and observation.

POMDPs.updater — Function.

updater(policy::Policy)

Returns a default Updater appropriate for a belief type that policy p can use

POMDPs.upperbound

POMDPs.value — Function.

value(p::Policy, s)
value(p::Policy, s, a)

Returns the utility value from policy p given the state (or belief), or state-action (or belief-action) pair.

The state-action version is commonly referred to as the Q-value.

POMDPs.verify_registry

POMDPs.weight