Docstrings

Rotations.AngleAxis — Type.

struct AngleAxis{T} <: Rotation{3,T}
AngleAxis(Θ, x, y, z)

A 3×3 rotation matrix parameterized by a 3D rotation by angle θ about an arbitrary axis [x, y, z].

Note that the axis is not unique for θ = 0, and that this parameterization does not continuously map the neighbourhood of the null rotation (and therefore might not be suitable for autodifferentation and optimization purposes).

Note: by default, the constructor will renormalize the input so that the axis has length 1 (x² + y² + z² = 1).

Renormalization can be skipped by passing false as an additional constructor argument, in which case the user provides the guarantee that the input arguments represent a normalized rotation axis. Operations on an AngleAxis with a rotation axis that does not have unit norm, created by skipping renormalization in this fashion, are not guaranteed to do anything sensible.

Rotations.Quat — Type.

struct Quat{T} <: Rotation{3,T}
Quat(w, x, y, z)

The Quat type is a 3×3 matrix representation of a normalized quaternion. They allow you to transparently use (fast) quaternion algebra to store, compose and invert 3D rotations, while at the same time letting you apply rotations through matrix-vector multiplication.

Note: by default, the constructor will renormalize the input so that the quaternion has length 1 (w² + x² + y² + z² = 1), and the rotation matrix is orthogonal.

Renormalization can be skipped by passing false as an additional constructor argument, in which case the user provides the guarantee that the input arguments represent a unit quaternion. Operations on an unnormalized Quat, created by skipping renormalization in this fashion, are not guaranteed to do anything sensible.

Rotations.RodriguesVec — Type.

struct RodriguesVec{T} <: Rotation{3,T}
RodriguesVec(sx, sy, sz)

Rodrigues vector parameterization of a 3×3 rotation matrix. The direction of the vector [sx, sy, sz] defines the axis of rotation, and the rotation angle is given by its norm.

Rotations.RotMatrix — Type.

struct RotMatrix{N,T} <: Rotation{N,T}

A statically-sized, N×N unitary (orthogonal) matrix.

Note: the orthonormality of the input matrix is not checked by the constructor.

Rotations.RotMatrix2

Rotations.RotMatrix3

Rotations.RotX — Type.

struct RotX{T} <: Rotation{3,T}
RotX(theta)

A 3×3 rotation matrix which represents a rotation by theta about the X axis.

Rotations.RotXY — Type.

struct RotXY{T} <: Rotation{3,T}
RotXY(theta_x, theta_y)

A 3×3 rotation matrix which represents a rotation by theta_y about the Y axis, followed by a rotation by theta_x about the X axis.

Rotations.RotXYX — Type.

struct RotXYX{T} <: Rotation{3,T}
RotXYX(theta1, theta2, theta3)

A 3×3 rotation matrix parameterized by the "proper" XYX Euler angle convention, consisting of first a rotation about the X axis by theta3, followed by a rotation about the Y axis by theta2, and finally a rotation about the X axis by theta1.

Rotations.RotXYZ — Type.

struct RotXYZ{T} <: Rotation{3,T}
RotXYZ(theta1, theta2, theta3)
RotXYZ(roll=r, pitch=p, yaw=y)

A 3×3 rotation matrix parameterized by the "Tait-Bryant" XYZ Euler angle convention, consisting of first a rotation about the Z axis by theta3, followed by a rotation about the Y axis by theta2, and finally a rotation about the X axis by theta1.

The keyword argument form applies roll, pitch and yaw to the X, Y and Z axes respectively, in XYZ order. (Because it is a right-handed coordinate system, note that positive pitch is heading in the negative Z axis).

Rotations.RotXZ — Type.

struct RotXZ{T} <: Rotation{3,T}
RotXZ(theta_x, theta_z)

A 3×3 rotation matrix which represents a rotation by theta_z about the Z axis, followed by a rotation by theta_x about the X axis.

Rotations.RotXZX — Type.

struct RotXZX{T} <: Rotation{3,T}
RotXZX(theta1, theta2, theta3)

A 3×3 rotation matrix parameterized by the "proper" XZX Euler angle convention, consisting of first a rotation about the X axis by theta3, followed by a rotation about the Z axis by theta2, and finally a rotation about the X axis by theta1.

Rotations.RotXZY — Type.

struct RotXZY{T} <: Rotation{3,T}
RotXZY(theta1, theta2, theta3)
RotXZY(roll=r, pitch=p, yaw=y)

A 3×3 rotation matrix parameterized by the "Tait-Bryant" XZY Euler angle convention, consisting of first a rotation about the Y axis by theta3, followed by a rotation about the Z axis by theta2, and finally a rotation about the X axis by theta1.

The keyword argument form applies roll, pitch and yaw to the X, Y and Z axes respectively, in XZY order. (Because it is a right-handed coordinate system, note that positive pitch is heading in the negative Z axis).

Rotations.RotY — Type.

struct RotY{T} <: Rotation{3,T}
RotY(theta)

A 3×3 rotation matrix which represents a rotation by theta about the Y axis.

Rotations.RotYX — Type.

struct RotYX{T} <: Rotation{3,T}
RotYX(theta_y, theta_x)

A 3×3 rotation matrix which represents a rotation by theta_x about the X axis, followed by a rotation by theta_y about the Y axis.

Rotations.RotYXY — Type.

struct RotYXY{T} <: Rotation{3,T}
RotYXY(theta1, theta2, theta3)

A 3×3 rotation matrix parameterized by the "proper" YXY Euler angle convention, consisting of first a rotation about the Y axis by theta3, followed by a rotation about the X axis by theta2, and finally a rotation about the Y axis by theta1.

Rotations.RotYXZ — Type.

struct RotYXZ{T} <: Rotation{3,T}
RotYXZ(theta1, theta2, theta3)
RotYXZ(roll=r, pitch=p, yaw=y)

A 3×3 rotation matrix parameterized by the "Tait-Bryant" YXZ Euler angle convention, consisting of first a rotation about the Z axis by theta3, followed by a rotation about the X axis by theta2, and finally a rotation about the Y axis by theta1.

The keyword argument form applies roll, pitch and yaw to the X, Y and Z axes respectively, in YXZ order. (Because it is a right-handed coordinate system, note that positive pitch is heading in the negative Z axis).

Rotations.RotYZ — Type.

struct RotYZ{T} <: Rotation{3,T}
RotYZ(theta_y, theta_z)

A 3×3 rotation matrix which represents a rotation by theta_z about the Z axis, followed by a rotation by theta_y about the Y axis.

Rotations.RotYZX — Type.

struct RotYZX{T} <: Rotation{3,T}
RotYZX(theta1, theta2, theta3)
RotYZX(roll=r, pitch=p, yaw=y)

A 3×3 rotation matrix parameterized by the "Tait-Bryant" YZX Euler angle convention, consisting of first a rotation about the X axis by theta3, followed by a rotation about the Z axis by theta2, and finally a rotation about the Y axis by theta1.

The keyword argument form applies roll, pitch and yaw to the X, Y and Z axes respectively, in YZX order. (Because it is a right-handed coordinate system, note that positive pitch is heading in the negative Z axis).

Rotations.RotYZY — Type.

struct RotYZY{T} <: Rotation{3,T}
RotYZY(theta1, theta2, theta3)

A 3×3 rotation matrix parameterized by the "proper" YXY Euler angle convention, consisting of first a rotation about the Y axis by theta3, followed by a rotation about the Z axis by theta2, and finally a rotation about the Y axis by theta1.

Rotations.RotZ — Type.

struct RotZ{T} <: Rotation{3,T}
RotZ(theta)

A 3×3 rotation matrix which represents a rotation by theta about the Z axis.

Rotations.RotZX — Type.

struct RotZX{T} <: Rotation{3,T}
RotZX(theta_z, theta_x)

A 3×3 rotation matrix which represents a rotation by theta_x about the X axis, followed by a rotation by theta_z about the Z axis.

Rotations.RotZXY — Type.

struct RotZXY{T} <: Rotation{3,T}
RotZXY(theta1, theta2, theta3)
RotZXY(roll=r, pitch=p, yaw=y)

A 3×3 rotation matrix parameterized by the "Tait-Bryant" ZXY Euler angle convention, consisting of first a rotation about the Y axis by theta3, followed by a rotation about the X axis by theta2, and finally a rotation about the Z axis by theta1.

The keyword argument form applies roll, pitch and yaw to the X, Y and Z axes respectively, in ZXY order. (Because it is a right-handed coordinate system, note that positive pitch is heading in the negative Z axis).

Rotations.RotZXZ — Type.

struct RotZXZ{T} <: Rotation{3,T}
RotZXZ(theta1, theta2, theta3)

A 3×3 rotation matrix parameterized by the "proper" ZXZ Euler angle convention, consisting of first a rotation about the Z axis by theta3, followed by a rotation about the X axis by theta2, and finally a rotation about the Z axis by theta1.

Rotations.RotZY — Type.

struct RotZY{T} <: Rotation{3,T}
RotZY(theta_z, theta_y)

A 3×3 rotation matrix which represents a rotation by theta_y about the Y axis, followed by a rotation by theta_z about the Z axis.

Rotations.RotZYX — Type.

struct RotZYX{T} <: Rotation{3,T}
RotZYX(theta1, theta2, theta3)
RotZYX(roll=r, pitch=p, yaw=y)

A 3×3 rotation matrix parameterized by the "Tait-Bryant" ZYX Euler angle convention, consisting of first a rotation about the X axis by theta3, followed by a rotation about the Y axis by theta2, and finally a rotation about the Z axis by theta1.

The keyword argument form applies roll, pitch and yaw to the X, Y and Z axes respectively, in ZYX order. (Because it is a right-handed coordinate system, note that positive pitch is heading in the negative Z axis).

Rotations.RotZYZ — Type.

struct RotZYZ{T} <: Rotation{3,T}
RotZYZ(theta1, theta2, theta3)

A 3×3 rotation matrix parameterized by the "proper" ZXZ Euler angle convention, consisting of first a rotation about the Z axis by theta3, followed by a rotation about the Y axis by theta2, and finally a rotation about the Z axis by theta1.

Rotations.Rotation — Type.

abstract type Rotation{N,T} <: StaticMatrix{N,N,T}

An abstract type representing N-dimensional rotations. More abstractly, they represent unitary (orthogonal) N×N matrices.

Rotations.Rotations

Rotations.SPQuat — Type.

struct SPQuat{T} <: Rotation{3,T}
SPQuat(x, y, z)

An SPQuat is a 3D rotation matrix represented by the "stereographic projection" of a normalized quaternion (shortened to "SPQuat"), which is a 3-element parametrization of a unit quaternion Q formed by the intersection of a line from [-1,0,0,0] to Q, with a plane containing the origin and with normal direction [1,0,0,0]. This is a compact representation of rotations where the derivitives of the rotation matrix's elements w.r.t. the SPQuat parameters are rational functions (making them useful for optimization).

See:

[1] "A Recipe on the Parameterization of Rotation Matrices for Non-Linear Optimization using Quaternions",
Terzakis, Culverhouse, Bugmann, Sharma, Sutton,
MIDAS technical report, 2011
http://www.tech.plymouth.ac.uk/sme/springerusv/2011/publications_files/Terzakis%20et%20al%202012,%20A%20Recipe%20on%20the%20Parameterization%20of%20Rotation%20Matrices...MIDAS.SME.2012.TR.004.pdf

Note 1: the singularity (origin) has been moved from [0,0,0,-1] in Ref[1] to [-1,0,0,0], so the 0 rotation quaternion [1,0,0,0] maps to [0,0,0] as opposed of to [1,0,0].
Note 2: the quaternion rotation ambiguity q = -q means there will be a rotation with ||SPQuat|| <= 1 and an equivilent rotation with ||SPQuat|| > 1.  This package
        uses the solution with ||SPQuat|| <= 1
Note 3: it is safe to assume that the corresponding matrix is orthogonal/unitary for any input x, y, z.

Rotations._sincos

Rotations.d_cross

Rotations.eval

Rotations.eye

Rotations.include

Rotations.isrotation — Function.

isrotation(r)
isrotation(r, tol)

Check whether r is a 3×3 rotation matrix, where r * r' is within tol of the identity matrix (using the Frobenius norm). (tol defaults to 1000 * eps(eltype(r)).)

Rotations.jacobian — Function.

jacobian(::Type{output_param}, R::input_param)

Returns the jacobian for transforming from the input rotation parameterization to the output parameterization, centered at the value of R.

jacobian(R::rotation_type, X::AbstractVector)

Returns the jacobian for rotating the vector X by R.

Rotations.perpendicular_vector — Function.

perpendicular_vector(vec)

Compute a vector perpendicular to vec by switching the two elements with largest absolute value, flipping the sign of the second largest, and setting the remaining element to zero.

Rotations.rotation_angle

Rotations.rotation_axis

Rotations.rotation_between — Function.

rotation_between(from, to)

Compute the quaternion that rotates vector from so that it aligns with vector to, along the geodesic (shortest path).