Docstrings

NLPModels.@lencheck

NLPModels.@rangecheck

NLPModels.ADNLPModel — Type.

ADNLPModel is an AbstractNLPModel using ForwardDiff to compute the derivatives. In this interface, the objective function $f$ and an initial estimate are required. If there are constraints, the function $c:\mathbb{R}^n\rightarrow\mathbb{R}^m$ and the vectors $c_L$ and $c_U$ also need to be passed. Bounds on the variables and an inital estimate to the Lagrangian multipliers can also be provided.

ADNLPModel(f, x0; lvar = [-∞,…,-∞], uvar = [∞,…,∞], y0 = zeros,
  c = NotImplemented, lcon = [-∞,…,-∞], ucon = [∞,…,∞], name = "Generic")

f :: Function - The objective function $f$;

x0 :: AbstractVector - The initial point of the problem;

lvar :: AbstractVector - $\ell$, the lower bound of the variables;

uvar :: AbstractVector - $u$, the upper bound of the variables;

c :: Function - The constraints function $c$;

y0 :: AbstractVector - The initial value of the Lagrangian estimates;

lcon :: AbstractVector - $c_L$, the lower bounds of the constraints function;

ucon :: AbstractVector - $c_U$, the upper bounds of the constraints function;

name :: String - A name for the model.

The functions follow the same restrictions of ForwardDiff functions, summarised here:

The function can only be composed of generic Julia functions;

The function must accept only one argument;

The function's argument must accept a subtype of AbstractVector;

The function should be type-stable.

For contrained problems, the function $c$ is required, and it must return an array even when m = 1, and $c_L$ and $c_U$ should be passed, otherwise the problem is ill-formed. For equality constraints, the corresponding index of $c_L$ and $c_U$ should be the same.

NLPModels.ADNLSModel — Type.

ADNLSModel is an Nonlinear Least Squares model using ForwardDiff to compute the derivatives.

ADNLSModel(F, x0, m; lvar = [-∞,…,-∞], uvar = [∞,…,∞], y0 = zeros,
  c = NotImplemented, lcon = [-∞,…,-∞], ucon = [∞,…,∞], name = "Generic")

F :: Function - The residual function $F$;

x0 :: AbstractVector - The initial point of the problem;

m :: Int - The dimension of $F(x)$, i.e., the number of

equations in the nonlinear system.

The other parameters are as in ADNLPModel.

NLPModels.AbstractNLPModel

NLPModels.AbstractNLPModelMeta

NLPModels.AbstractNLSModel

NLPModels.Counters

NLPModels.FeasibilityFormNLS — Type.

Converts a nonlinear least-squares problem with residual F(x) to a nonlinear optimization problem with constraints F(x) = r and objective ¹/₂‖r‖². In other words, converts

min ¹/₂‖F(x)‖²
s.t  cₗ ≤ c(x) ≤ cᵤ
      ℓ ≤   x  ≤ u

to

min ¹/₂‖r‖²
s.t   F(x) - r = 0
     cₗ ≤ c(x) ≤ cᵤ
      ℓ ≤   x  ≤ u

If you rather have the first problem, the nls model already works as an NLPModel of that format.

NLPModels.FeasibilityResidual — Type.

A feasibility residual model is created from a NLPModel of the form

min f(x)
s.t cℓ ≤ c(x) ≤ cu
    bℓ ≤   x  ≤ bu

by creating slack variables s and defining F(x,s) = c(x) - s. The resulting NLS problem is

min ¹/₂‖c(x) - s‖²
    bℓ ≤ x ≤ bu
    cℓ ≤ s ≤ bu

This is done using SlackModel first, and then defining the NLS. Notice that if bℓᵢ = buᵢ, no slack variable is created.

NLPModels.LBFGSModel — Type.

Construct a LBFGSModel from another type of model.

NLPModels.LLSModel — Type.

nls = LLSModel(A, b; lvar, uvar, C, lcon, ucon)

Creates a Linear Least Squares model ½‖Ax - b‖² with optional bounds lvar ≦ x ≦ y and optional linear constraints lcon ≦ Cx ≦ ucon.

NLPModels.LSR1Model — Type.

Construct a LSR1Model from another type of nlp.

NLPModels.NLPModelMeta

NLPModels.NLPModels

NLPModels.NLSCounters

NLPModels.NLSMeta

NLPModels.NotImplemented

NLPModels.NotImplementedError

NLPModels.QuasiNewtonModel

NLPModels.SimpleNLPModel — Type.

SimpleNLPModel is an AbstractNLPModel that uses only user-defined functions. In this interface, the objective function $f$ and an initial estimate are required. If the user wants to use derivatives, they need to be passed. The same goes for the Hessian and Hessian-AbstractVector product. For constraints, $c:\mathbb{R}^n\rightarrow\mathbb{R}^m$ and the vectors $c_L$ and $c_U$ also need to be passed. Bounds on the variables and an inital estimate to the Lagrangian multipliers can also be provided. The user can also pass the Jacobian and the Lagrangian Hessian and Hessian-AbstractVector product.

SimpleNLPModel(f, x0; lvar = [-∞,…,-∞], uvar = [∞,…,∞], y0=zeros,
  lcon = [-∞,…,-∞], ucon = [∞,…,∞], name = "Generic",
  [list of functions])

f :: Function - The objective function $f$;

x0 :: AbstractVector - The initial point of the problem;

lvar :: AbstractVector - $\ell$, the lower bound of the variables;

uvar :: AbstractVector - $u$, the upper bound of the variables;

y0 :: AbstractVector - The initial value of the Lagrangian estimates;

lcon :: AbstractVector - $c_L$, the lower bounds of the constraints function;

ucon :: AbstractVector - $c_U$, the upper bounds of the constraints function;

name :: String - A name for the model.

All functions passed have a direct correlation with a NLP function. You don't have to define any more than you need, but calling an undefined function will throw a NotImplementedError. The list is

g and g!: $\nabla f(x)$, the gradient of the objective function;

gx = g(x)
gx = g!(x, gx)

H: The lower triangle of the Hessian of the objective function or of the Lagrangian;

Hx = H(x; obj_weight=1.0) # if the problem is unconstrained
Hx = H(x; obj_weight=1.0, y=zeros) # if the problem is constrained

Hcoord - The lower triangle of the Hessian of the objective function or of the Lagrangian, in triplet format;

(rows,cols,vals) = Hcoord(x; obj_weight=1.0) # if the problem is unconstrained
(rows,cols,vals) = Hcoord(x; obj_weight=1.0, y=zeros) # if the problem is constrained

Hp and Hp! - The product of the Hessian of the objective function or of the Lagrangian by a vector;

Hv = Hp(x, v, obj_weight=1.0) # if the problem is unconstrained
Hv = Hp!(x, v, Hv, obj_weight=1.0) # if the problem is unconstrained
Hv = Hp(x, v, obj_weight=1.0, y=zeros) # if the problem is constrained
Hv = Hp!(x, v, Hv, obj_weight=1.0, y=zeros) # if the problem is constrained

c and c! - $c(x)$, the constraints function;

cx = c(x)
cx = c!(x, cx)

J - $J(x)$, the Jacobian of the constraints;

Jx = J(x)

Jcoord - $J(x)$, the Jacobian of the constraints, in triplet format;

(rows,cols,vals) = Jcoord(x)

Jp and Jp! - The Jacobian-vector product;

Jv = Jp(x, v)
Jv = Jp!(x, v, Jv)

Jtp and Jtp! - The Jacobian-transposed-vector product;

Jtv = Jtp(x, v)
Jtv = Jtp!(x, v, Jtv)

For contrained problems, the function $c$ is required, and it must return an array even when m = 1, and $c_L$ and $c_U$ should be passed, otherwise the problem is ill-formed. For equality constraints, the corresponding index of $c_L$ and $c_U$ should be the same.

NLPModels.SimpleNLSModel — Type.

nls = SimpleNLSModel(n;  F=F, F! =F!, JF=JF, JFp=JFp, JFp! =JFp!,
JFtp=JFtp, JFtp! =JFtp!)
nls = SimpleNLSModel(x0; F=F, F! =F!, JF=JF, JFp=JFp, JFp! =JFp!,
JFtp=JFtp, JFtp! =JFtp!)

Creates a Nonlinear Linear Least Squares model to minimize ‖F(x)‖². If JF = JF(x) is passed, the Jacobian is available.

NLPModels.SlackModel — Type.

A model whose only inequality constraints are bounds.

Given a model, this type represents a second model in which slack variables are introduced so as to convert linear and nonlinear inequality constraints to equality constraints and bounds. More precisely, if the original model has the form

\[ \min f(x) \mbox{ s. t. } cL \leq c(x) \leq cU \mbox{ and } \ell \leq x \leq u, \]

the new model appears to the user as

\[ \min f(X) \mbox{ s. t. } g(X) = 0 \mbox{ and } L \leq X \leq U. \]

The unknowns $X = (x, s)$ contain the original variables and slack variables $s$. The latter are such that the new model has the general form

\[ \min f(x) \mbox{ s. t. } c(x) - s = 0, cL \leq s \leq cU \mbox{ and } \ell \leq x \leq u, \]

although no slack variables are introduced for equality constraints.

The slack variables are implicitly ordered as [s(low), s(upp), s(rng)], where low, upp and rng represent the indices of the constraints of the form $c_L \leq c(x) < \infty$, $-\infty < c(x) \leq c_U$ and $c_L \leq c(x) \leq c_U$, respectively.

NLPModels.SlackModels

NLPModels.SlackNLSModel — Type.

Like SlackModel, this model converts inequalities into equalities and bounds.

NLPModels.bound_constrained — Function.

bound_constrained(nlp)
bound_constrained(meta)

Returns whether the problem has bounds on the variables and no other constraints.

NLPModels.cons — Function.

c = cons(nlp, x)

Evaluate $c(x)$, the constraints at x.

NLPModels.cons! — Function.

c = cons!(nlp, x, c)

Evaluate $c(x)$, the constraints at x in place.

NLPModels.conscale

NLPModels.counters

NLPModels.equality_constrained — Function.

equality_constrained(nlp)
equality_constrained(meta)

Returns whether the problem's constraints are all equalities. Unconstrained problems return false.

NLPModels.eval

NLPModels.ghjvprod

NLPModels.ghjvprod!

NLPModels.grad — Function.

g = grad(nlp, x)

Evaluate $\nabla f(x)$, the gradient of the objective function at x.

NLPModels.grad! — Function.

g = grad!(nlp, x, g)

Evaluate $\nabla f(x)$, the gradient of the objective function at x in place.

NLPModels.gradient_check — Function.

Check the first derivatives of the objective at x against centered finite differences.

This function returns a dictionary indexed by components of the gradient for which the relative error exceeds rtol.

NLPModels.has_bounds — Function.

has_bounds(nlp)
has_bounds(meta)

Returns whether the problem has bounds on the variables.

NLPModels.hess — Function.

Hx = hess(nlp, x; obj_weight=1.0, y=zeros)

Evaluate the Lagrangian Hessian at (x,y) as a sparse matrix, with objective function scaled by obj_weight, i.e.,

\[ \nabla^2L(x,y) = \sigma * \nabla^2 f(x) + \sum{i=1}^m yi\nabla^2 c_i(x), \]

with σ = obj_weight. Only the lower triangle is returned.

NLPModels.hess_coord — Function.

(rows,cols,vals) = hess_coord(nlp, x; obj_weight=1.0, y=zeros)

Evaluate the Lagrangian Hessian at (x,y) in sparse coordinate format, with objective function scaled by obj_weight, i.e.,

\[ \nabla^2L(x,y) = \sigma * \nabla^2 f(x) + \sum{i=1}^m yi\nabla^2 c_i(x), \]

with σ = obj_weight. Only the lower triangle is returned.

NLPModels.hess_coord_residual — Function.

(rows,cols,vals) = hess_coord_residual(nls, x, v)

Computes the linear combination of the Hessians of the residuals at x with coefficients v in sparse coordinate format.

NLPModels.hess_op — Function.

H = hess_op(nlp, x; obj_weight=1.0, y=zeros)

Return the Lagrangian Hessian at (x,y) with objective function scaled by obj_weight as a linear operator. The resulting object may be used as if it were a matrix, e.g., H * v. The linear operator H represents

\[ \nabla^2L(x,y) = \sigma * \nabla^2 f(x) + \sum{i=1}^m yi\nabla^2 c_i(x), \]

with σ = obj_weight.

NLPModels.hess_op! — Function.

H = hess_op!(nlp, x, Hv; obj_weight=1.0, y=zeros)

Return the Lagrangian Hessian at (x,y) with objective function scaled by obj_weight as a linear operator, and storing the result on Hv. The resulting object may be used as if it were a matrix, e.g., w = H * v. The vector Hv is used as preallocated storage for the operation. The linear operator H represents

\[ \nabla^2L(x,y) = \sigma * \nabla^2 f(x) + \sum{i=1}^m yi\nabla^2 c_i(x), \]

with σ = obj_weight.

NLPModels.hess_op_residual — Function.

Hop = hess_op_residual(nls, x, i)

Computes the Hessian of the i-th residual at x, in linear operator form.

NLPModels.hess_op_residual! — Function.

Hop = hess_op_residual!(nls, x, i, Hiv)

Computes the Hessian of the i-th residual at x, in linear operator form. The vector Hiv is used as preallocated storage for the operation.

NLPModels.hess_residual — Function.

H = hess_residual(nls, x, v)

Computes the linear combination of the Hessians of the residuals at x with coefficients v.

NLPModels.hessian_check — Function.

Check the second derivatives of the objective and each constraints at x against centered finite differences. This check does not rely on exactness of the first derivatives, only on objective and constraint values.

The sgn arguments refers to the formulation of the Lagrangian in the problem. It should have a positive value if the Lagrangian is formulated as

L(x,y) = f(x) + ∑ yⱼ cⱼ(x)

e.g., as in JuMPNLPModels, and a negative value if the Lagrangian is formulated as

L(x,y) = f(x) - ∑ yⱼ cⱼ(x)

e.g., as in AmplModels. Only the sign of sgn is important.

This function returns a dictionary indexed by functions. The 0-th function is the objective while the k-th function (for k > 0) is the k-th constraint. The values of the dictionary are dictionaries indexed by tuples (i, j) such that the relative error in the second derivative ∂²fₖ/∂xᵢ∂xⱼ exceeds rtol.

NLPModels.hessian_check_from_grad — Function.

Check the second derivatives of the objective and each constraints at x against centered finite differences. This check assumes exactness of the first derivatives.

The sgn arguments refers to the formulation of the Lagrangian in the problem. It should have a positive value if the Lagrangian is formulated as

L(x,y) = f(x) + ∑ yⱼ cⱼ(x)

e.g., as in JuMPNLPModels, and a negative value if the Lagrangian is formulated as

L(x,y) = f(x) - ∑ yⱼ cⱼ(x)

e.g., as in AmplModels. Only the sign of sgn is important.

This function returns a dictionary indexed by functions. The 0-th function is the objective while the k-th function (for k > 0) is the k-th constraint. The values of the dictionary are dictionaries indexed by tuples (i, j) such that the relative error in the second derivative ∂²fₖ/∂xᵢ∂xⱼ exceeds rtol.

NLPModels.hprod — Function.

Hv = hprod(nlp, x, v; obj_weight=1.0, y=zeros)

Evaluate the product of the Lagrangian Hessian at (x,y) with the vector v, with objective function scaled by obj_weight, i.e.,

\[ \nabla^2L(x,y) = \sigma * \nabla^2 f(x) + \sum{i=1}^m yi\nabla^2 c_i(x), \]

with σ = obj_weight.

NLPModels.hprod! — Function.

Hv = hprod!(nlp, x, v, Hv; obj_weight=1.0, y=zeros)

Evaluate the product of the Lagrangian Hessian at (x,y) with the vector v in place, with objective function scaled by obj_weight, i.e.,

\[ \nabla^2L(x,y) = \sigma * \nabla^2 f(x) + \sum{i=1}^m yi\nabla^2 c_i(x), \]

with σ = obj_weight.

NLPModels.hprod_residual — Function.

Hiv = hprod_residual(nls, x, i, v)

Computes the product of the Hessian of the i-th residual at x, times the vector v.

NLPModels.hprod_residual! — Function.

Hiv = hprod_residual!(nls, x, i, v, Hiv)

Computes the product of the Hessian of the i-th residual at x, times the vector v, and stores it in vector Hiv.

NLPModels.include

NLPModels.increment! — Function.

increment!(nlp, s)

Increment counter s of problem nlp.

NLPModels.inequality_constrained — Function.

inequality_constrained(nlp)
inequality_constrained(meta)

Returns whether the problem's constraints are all inequalities. Unconstrained problems return true.

NLPModels.jac — Function.

Jx = jac(nlp, x)

Evaluate $\nabla c(x)$, the constraint's Jacobian at x as a sparse matrix.

NLPModels.jac_coord — Function.

(rows,cols,vals) = jac_coord(nlp, x)

Evaluate $\nabla c(x)$, the constraint's Jacobian at x in sparse coordinate format.

NLPModels.jac_coord_residual — Function.

(rows,cols,vals) = jac_coord_residual(nls, x)

Computes the Jacobian of the residual at x in sparse coordinate format.

NLPModels.jac_op — Function.

J = jac_op(nlp, x)

Return the Jacobian at x as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v.

NLPModels.jac_op! — Function.

J = jac_op!(nlp, x, Jv, Jtv)

Return the Jacobian at x as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v. The values Jv and Jtv are used as preallocated storage for the operations.

NLPModels.jac_op_residual — Function.

Jx = jac_op_residual(nls, x)

Computes J(x), the Jacobian of the residual at x, in linear operator form.

NLPModels.jac_op_residual! — Function.

Jx = jac_op_residual!(nls, x, Jv, Jtv)

Computes J(x), the Jacobian of the residual at x, in linear operator form. The vectors Jv and Jtv are used as preallocated storage for the operations.

NLPModels.jac_residual — Function.

Jx = jac_residual(nls, x)

Computes J(x), the Jacobian of the residual at x.

NLPModels.jacobian_check — Function.

Check the first derivatives of the constraints at x against centered finite differences.

This function returns a dictionary indexed by (j, i) tuples such that the relative error in the i-th partial derivative of the j-th constraint exceeds rtol.

NLPModels.jprod — Function.

Jv = jprod(nlp, x, v)

Evaluate $\nabla c(x)v$, the Jacobian-vector product at x.

NLPModels.jprod! — Function.

Jv = jprod!(nlp, x, v, Jv)

Evaluate $\nabla c(x)v$, the Jacobian-vector product at x in place.

NLPModels.jprod_residual — Function.

Jv = jprod_residual(nls, x, v)

Computes the product of the Jacobian of the residual at x and a vector, i.e., J(x)*v.

NLPModels.jprod_residual! — Function.

Jv = jprod_residual!(nls, x, v, Jv)

Computes the product of the Jacobian of the residual at x and a vector, i.e., J(x)*v, storing it in Jv.

NLPModels.jth_con

NLPModels.jth_congrad

NLPModels.jth_congrad!

NLPModels.jth_hess_residual — Function.

Hj = jth_hess_residual(nls, x, j)

Computes the Hessian of the j-th residual at x.

NLPModels.jth_hprod

NLPModels.jth_hprod!

NLPModels.jth_sparse_congrad

NLPModels.jtprod — Function.

Jtv = jtprod(nlp, x, v, Jtv)

Evaluate $\nabla c(x)^Tv$, the transposed-Jacobian-vector product at x.

NLPModels.jtprod! — Function.

Jtv = jtprod!(nlp, x, v, Jtv)

Evaluate $\nabla c(x)^Tv$, the transposed-Jacobian-vector product at x in place.

NLPModels.jtprod_residual — Function.

Jtv = jtprod_residual(nls, x, v)

Computes the product of the transpose of the Jacobian of the residual at x and a vector, i.e., J(x)'*v.

NLPModels.jtprod_residual! — Function.

Jtv = jtprod_residual!(nls, x, v, Jtv)

Computes the product of the transpose of the Jacobian of the residual at x and a vector, i.e., J(x)'*v, storing it in Jtv.

NLPModels.lagscale

NLPModels.linearly_constrained — Function.

linearly_constrained(nlp)
linearly_constrained(meta)

Returns whether the problem's constraints are known to be all linear.

NLPModels.neval_cons — Function.

NLPModels.neval_cons(nlp)

Get the number of cons evaluations.

NLPModels.neval_grad — Function.

NLPModels.neval_grad(nlp)

Get the number of grad evaluations.

NLPModels.neval_hess — Function.

NLPModels.neval_hess(nlp)

Get the number of hess evaluations.

NLPModels.neval_hess_residual — Function.

NLPModels.neval_hess_residual(nlp)

Get the number of hess evaluations.

NLPModels.neval_hprod — Function.

NLPModels.neval_hprod(nlp)

Get the number of hprod evaluations.

NLPModels.neval_hprod_residual — Function.

NLPModels.neval_hprod_residual(nlp)

Get the number of hprod evaluations.

NLPModels.neval_jac — Function.

NLPModels.neval_jac(nlp)

Get the number of jac evaluations.

NLPModels.neval_jac_residual — Function.

NLPModels.neval_jac_residual(nlp)

Get the number of jac evaluations.

NLPModels.neval_jcon — Function.

NLPModels.neval_jcon(nlp)

Get the number of jcon evaluations.

NLPModels.neval_jgrad — Function.

NLPModels.neval_jgrad(nlp)

Get the number of jgrad evaluations.

NLPModels.neval_jhess_residual — Function.

NLPModels.neval_jhess_residual(nlp)

Get the number of jhess evaluations.

NLPModels.neval_jhprod — Function.

NLPModels.neval_jhprod(nlp)

Get the number of jhprod evaluations.

NLPModels.neval_jprod — Function.

NLPModels.neval_jprod(nlp)

Get the number of jprod evaluations.

NLPModels.neval_jprod_residual — Function.

NLPModels.neval_jprod_residual(nlp)

Get the number of jprod evaluations.

NLPModels.neval_jtprod — Function.

NLPModels.neval_jtprod(nlp)

Get the number of jtprod evaluations.

NLPModels.neval_jtprod_residual — Function.

NLPModels.neval_jtprod_residual(nlp)

Get the number of jtprod evaluations.

NLPModels.neval_obj — Function.

NLPModels.neval_obj(nlp)

Get the number of obj evaluations.

NLPModels.neval_residual — Function.

NLPModels.neval_residual(nlp)

Get the number of residual evaluations.

NLPModels.nls_meta

NLPModels.obj — Function.

f = obj(nlp, x)

Evaluate $f(x)$, the objective function of nlp at x.

NLPModels.objcons — Function.

f, c = objcons(nlp, x)

Evaluate $f(x)$ and $c(x)$ at x.

NLPModels.objcons! — Function.

f = objcons!(nlp, x, c)

Evaluate $f(x)$ and $c(x)$ at x. c is overwritten with the value of $c(x)$.

NLPModels.objgrad — Function.

f, g = objgrad(nlp, x)

Evaluate $f(x)$ and $\nabla f(x)$ at x.

NLPModels.objgrad! — Function.

f, g = objgrad!(nlp, x, g)

Evaluate $f(x)$ and $\nabla f(x)$ at x. g is overwritten with the value of $\nabla f(x)$.

NLPModels.push!

LinearOperators.reset! — Function.

reset!(counters)

Reset evaluation counters

`reset!(nlp)

Reset evaluation count in nlp

NLPModels.residual — Function.

Fx = residual(nls, x)

Computes F(x), the residual at x.

NLPModels.residual! — Function.

Fx = residual!(nls, x, Fx)

Computes F(x), the residual at x.

NLPModels.slack_meta

NLPModels.sum_counters — Function.

sum_counters(counters)

Sum all counters of counters.

sum_counters(nlp)

Sum all counters of problem nlp.

NLPModels.unconstrained — Function.

unconstrained(nlp)
unconstrained(meta)

Returns whether the problem in unconstrained.

NLPModels.varscale