Docstrings

JavaCall.@jimport

JavaCall.JAVA_HOME_CANDIDATES

JavaCall.JClass

JavaCall.JClassLoader

JavaCall.JMethod

JavaCall.JNIEnv

JavaCall.JNIInvokeInterface

JavaCall.JNINativeInterface

JavaCall.JNI_EDETACHED

JavaCall.JNI_EEXIST

JavaCall.JNI_EINVAL

JavaCall.JNI_ENOMEM

JavaCall.JNI_ERR

JavaCall.JNI_EVERSION

JavaCall.JNI_FALSE

JavaCall.JNI_OK

JavaCall.JNI_TRUE

JavaCall.JNI_VERSION_1_1

JavaCall.JNI_VERSION_1_2

JavaCall.JNI_VERSION_1_4

JavaCall.JNI_VERSION_1_6

JavaCall.JNI_VERSION_1_8

JavaCall.JObject

JavaCall.JString

JavaCall.JThread

JavaCall.JavaCall

JavaCall.JavaCallError

JavaCall.JavaMetaClass

JavaCall.JavaObject

JavaCall.JavaVM

JavaCall.JavaVMInitArgs

JavaCall.JavaVMOption

JavaCall.__init__

JavaCall._jcall

JavaCall._jfield

JavaCall._jimport

JavaCall._jmc_cache

JavaCall._metaclass

JavaCall.addClassPath

JavaCall.addOpts

JavaCall.assertloaded

JavaCall.assertnotloaded

JavaCall.classforname — Function.

classforname(name::String)

Create an instance of Class<name> (same as Class.forName(name) in Java)

Args

name: The name of a class to instantiate

Returns

JavaObject Instance of Class<name>

JavaCall.conventional_name

JavaCall.convert_arg

JavaCall.convert_args

JavaCall.convert_result

JavaCall.cp

JavaCall.create

JavaCall.deleteref

JavaCall.destroy

JavaCall.eval

JavaCall.findjvm

JavaCall.getclass

JavaCall.geterror

JavaCall.getname — Function.

getname(cls::JClass)

Returns the fully qualified name of the java class

Args

cls: The JClass object

Returns

The fully qualified name of the java class

getname(method::JMethod)

Returns the fully qualified name of the java method

Args

cls: The JClass method

Returns

The fully qualified name of the method

JavaCall.getparametertypes — Function.

getparametertypes(method::JMethod)

Returns the parameter types of the java method

Args

method: The java method object

Returns

Vector the parametertypes

JavaCall.getreturntype — Function.

getreturntype(method::JMethod)

Returns the return type of the java method

Args

method: The java method object

Returns

Returns the type of the return object as a JClass

JavaCall.has_next

JavaCall.include

JavaCall.init

JavaCall.isConvertible

JavaCall.isloaded

JavaCall.isnull — Function.

isnull(obj::JavaObject)

Checks if the passed JavaObject is null or not

Args

obj: The object of type JavaObject

Returns

true if the passed object is null else false

isnull(obj::JavaMetaClass)

Checks if the passed JavaMetaClass is null or not

Args

obj: The object of type JavaMetaClass

Returns

true if the passed object is null else false

JavaCall.iterator

JavaCall.javaclassname

JavaCall.javahome_winreg

JavaCall.jboolean

JavaCall.jbyte

JavaCall.jcall

JavaCall.jchar

JavaCall.jdouble

JavaCall.jfield

JavaCall.jfloat

JavaCall.jint

JavaCall.jlong

JavaCall.jnew

JavaCall.jnifunc

JavaCall.jprimitive

JavaCall.jshort

JavaCall.jsize

JavaCall.jvalue

JavaCall.jvmfunc

JavaCall.libjvm

JavaCall.libname

JavaCall.listmethods — Function.

listmethods(obj::JavaObject)

Lists the methods that are available on the java object passed

Args

obj: The java object

Returns

List of methods

listmethods(obj::JavaObject, name::AbstractString)

Lists the methods that are available on the java object passed. The methods are filtered based on the name passed

Args

obj: The java object

name: The filter (e.g. method name)

Returns

List of methods available on the java object and matching the name passed

JavaCall.metaclass

JavaCall.method_signature

JavaCall.narrow — Function.

Given a JavaObject{T} narrows down T to a real class of the underlying object. For example, JavaObject{:java.lang.Object} pointing to java.lang.String will be narrowed down to JavaObject{:java.lang.String}

JavaCall.opts

JavaCall.penv

JavaCall.pjvm

JavaCall.real_jtype

JavaCall.sep

JavaCall.signature