Docstrings

FixedEffectModels.@formula

FixedEffectModels.@model — Macro.

Capture and parse a set of expressions to generate a Model

Arguments

ex::Expr: an expression parsed as a Formula

fe::Expr: Fixed effect expression. You can add an arbitrary number of high dimensional fixed effects, separated with +. Interact multiple categorical variables using & . Interact a categorical variable with a continuous variable using &. Alternative, use * to add a categorical variable and its interaction with a continuous variable. Variables must be of type CategoricalArray (use categorical to convert a variable to a CategoricalArray).

vcov::Expr: Vcov formula. Default to simple. robust and cluster() are also implemented

weights::Expr: Weight variable. Corresponds to analytical weights

subset::Expr: Expression of the form State .>= 30

Returns

::Model: a Model struct

Detail

A typical formula is composed of one dependent variable, exogeneous variables, endogeneous variables, and instruments

depvar ~ exogeneousvars + (endogeneousvars ~ instrumentvars

Examples

using DataFrames, RDatasets, FixedEffectModels
df = dataset("plm", "Cigar")
df[:StateC] =  categorical(df[:State])
df[:YearC] =  categorical(df[:Year])
reg(df, @model(Sales ~ NDI, weights = Pop))
@model(Sales ~ NDI, fe = StateC, vcov = robust)
@model(Sales ~ NDI, fe = StateC + YearC, weights = Pop, vcov = cluster(StateC)

FixedEffectModels.AbstractContrasts

FixedEffectModels.AbstractHistogram

FixedEffectModels.AbstractRegressionResult

FixedEffectModels.AbstractVcovFormula

FixedEffectModels.AbstractVcovMethod

FixedEffectModels.AbstractWeights

FixedEffectModels.AnalyticWeights

FixedEffectModels.CoefTable

FixedEffectModels.CoefTable2

FixedEffectModels.Combination

FixedEffectModels.ContrastsCoding

FixedEffectModels.ConvergenceException

FixedEffectModels.DummyCoding

FixedEffectModels.ECDF

FixedEffectModels.EffectsCoding

FixedEffectModels.FixedEffect

FixedEffectModels.FixedEffectMatrix

FixedEffectModels.FixedEffectModels

FixedEffectModels.FixedEffects

FixedEffectModels.Formula

FixedEffectModels.FrequencyWeights

FixedEffectModels.HelmertCoding

FixedEffectModels.Histogram

FixedEffectModels.L1dist

FixedEffectModels.L2dist

FixedEffectModels.Linfdist

FixedEffectModels.Model

FixedEffectModels.ModelFrame

FixedEffectModels.ModelFrame2

FixedEffectModels.ModelMatrix

FixedEffectModels.ProbabilityWeights

FixedEffectModels.RegressionModel

FixedEffectModels.RegressionResult

FixedEffectModels.RegressionResultFE

FixedEffectModels.RegressionResultFEIV

FixedEffectModels.RegressionResultIV

FixedEffectModels.StatisticalModel

FixedEffectModels.StatsBase

FixedEffectModels.StatsModels

FixedEffectModels.VcovClusterFormula

FixedEffectModels.VcovClusterMethod

FixedEffectModels.VcovData

FixedEffectModels.VcovDataMatrix

FixedEffectModels.VcovDataVector

FixedEffectModels.VcovFormula

FixedEffectModels.VcovMethod

FixedEffectModels.VcovRobustFormula

FixedEffectModels.VcovRobustMethod

FixedEffectModels.VcovSimpleFormula

FixedEffectModels.VcovSimpleMethod

FixedEffectModels.VcovWhiteMethod

FixedEffectModels.WeightFormula

FixedEffectModels.Weights

FixedEffectModels._check

FixedEffectModels._multiply

FixedEffectModels._multiply!

FixedEffectModels._name

FixedEffectModels._split

FixedEffectModels._subset

FixedEffectModels.addcounts!

FixedEffectModels.adjr2

FixedEffectModels.adjr²

FixedEffectModels.aic

FixedEffectModels.aicc

FixedEffectModels.allvars

FixedEffectModels.autocor

FixedEffectModels.autocor!

FixedEffectModels.autocov

FixedEffectModels.autocov!

FixedEffectModels.aweights

FixedEffectModels.basecol

FixedEffectModels.bic

FixedEffectModels.check

FixedEffectModels.coef

FixedEffectModels.coefnames

FixedEffectModels.coeftable

FixedEffectModels.competerank

FixedEffectModels.compute_Fstat

FixedEffectModels.compute_tss

FixedEffectModels.confint

FixedEffectModels.cor2cov

FixedEffectModels.corkendall

FixedEffectModels.corspearman

FixedEffectModels.counteq

FixedEffectModels.countmap

FixedEffectModels.countne

FixedEffectModels.counts

FixedEffectModels.cov2cor

FixedEffectModels.crosscor

FixedEffectModels.crosscor!

FixedEffectModels.crosscov

FixedEffectModels.crosscov!

FixedEffectModels.crossentropy

FixedEffectModels.crossprod

FixedEffectModels.decompose_iv!

FixedEffectModels.denserank

FixedEffectModels.describe

FixedEffectModels.deviance

FixedEffectModels.df_FStat

FixedEffectModels.df_residual

FixedEffectModels.dof

FixedEffectModels.dof_residual

FixedEffectModels.dropresponse

FixedEffectModels.dropterm

FixedEffectModels.ecdf

FixedEffectModels.entropy

FixedEffectModels.eval

FixedEffectModels.evaluate_subset

FixedEffectModels.fes

FixedEffectModels.findat

FixedEffectModels.fit

FixedEffectModels.fit!

FixedEffectModels.fitted

FixedEffectModels.format_scientific

FixedEffectModels.fweights

FixedEffectModels.genmean

FixedEffectModels.genvar

FixedEffectModels.geomean

FixedEffectModels.get_weights

FixedEffectModels.getcols

FixedEffectModels.gkldiv

FixedEffectModels.group

FixedEffectModels.harmmean

FixedEffectModels.helper_cluster

FixedEffectModels.hist

FixedEffectModels.include

FixedEffectModels.indexmap

FixedEffectModels.indicatormat

FixedEffectModels.informationmatrix

FixedEffectModels.inverse_rle

FixedEffectModels.iqr

FixedEffectModels.is_iv

FixedEffectModels.isfitted

FixedEffectModels.islinear

FixedEffectModels.isnaorneg

FixedEffectModels.isnested

FixedEffectModels.kldivergence

FixedEffectModels.kurtosis

FixedEffectModels.levelsmap

FixedEffectModels.leverage

FixedEffectModels.loglikelihood

FixedEffectModels.mad

FixedEffectModels.mad!

FixedEffectModels.maxad

FixedEffectModels.mean_and_cov

FixedEffectModels.mean_and_std

FixedEffectModels.mean_and_var

FixedEffectModels.meanad

FixedEffectModels.meanresponse

FixedEffectModels.middle

FixedEffectModels.midpoints

FixedEffectModels.mode

FixedEffectModels.model_helper

FixedEffectModels.model_response

FixedEffectModels.modelmatrix

FixedEffectModels.modes

FixedEffectModels.moment

FixedEffectModels.msd

FixedEffectModels.mss

FixedEffectModels.nobs

FixedEffectModels.nonmissing

FixedEffectModels.norepeats

FixedEffectModels.nquantile

FixedEffectModels.nulldeviance

FixedEffectModels.nullloglikelihood

FixedEffectModels.ordinalrank

FixedEffectModels.pacf

FixedEffectModels.pacf!

FixedEffectModels.parse_fixedeffect

FixedEffectModels.partial_out — Function.

Partial out variables in a Dataframe

Arguments

df::AbstractDataFrame

model::Model: A Model created using @model. See @model.

add_mean::Bool: Should the initial mean added to the returned variable?

method::Symbol: A symbol for the method. Default is :lsmr (akin to conjugate gradient descent). Other choices are :qr and :cholesky (factorization methods)

maxiter::Integer: Maximum number of iterations

tol::Real: Tolerance

Returns

::DataFrame: a dataframe with as many columns as there are dependent variables and as many rows as the original dataframe.

::Vector{Int}: a vector of iterations for each column

::Vector{Bool}: a vector of success for each column

Details

partial_out returns the residuals of a set of variables after regressing them on a set of regressors. The syntax is similar to reg - but it accepts multiple dependent variables. It returns a dataframe with as many columns as there are dependent variables and as many rows as the original dataframe. The regression model is estimated on only the rows where none of the dependent variables is missing. With the option add_mean = true, the mean of the initial variable is added to the residuals.

Examples

using  RDatasets, DataFrames, FixedEffectModels, Gadfly
df = dataset("datasets", "iris")
df[:SpeciesC] =  categorical(df[:Species])
result = partial_out(df, @model(SepalWidth + SepalLength ~ 1, fe = SpeciesC), add_mean = true)
plot(layer(result[1], x="SepalWidth", y="SepalLength", Stat.binmean(n=10), Geom.point),
   layer(result[1], x="SepalWidth", y="SepalLength", Geom.smooth(method=:lm)))

FixedEffectModels.partialcor

FixedEffectModels.percentile

FixedEffectModels.pinvertible

FixedEffectModels.predict

FixedEffectModels.predict!

FixedEffectModels.proportionmap

FixedEffectModels.proportions

FixedEffectModels.psnr

FixedEffectModels.pweights

FixedEffectModels.r2

FixedEffectModels.ranktest!

FixedEffectModels.reg — Function.

Estimate a linear model with high dimensional categorical variables / instrumental variables

Arguments

df::AbstractDataFrame

model::Model: A model created using @model

save::Union{Bool, Symbol} = false: Should residuals and eventual estimated fixed effects saved in a dataframe? Use save = :residuals to only save residuals. Use save = :fe to only save fixed effects.

method::Symbol = :lsmr: Method to deman regressors. :lsmr is akin to conjugate gradient descent. With parallel use :lsmr_parallel. To use multi threaded use lsmr_threads. Other choices are :qr and :cholesky (factorization methods)

contrasts::Dict = Dict() An optional Dict of contrast codings for each categorical variable in the formula. Any unspecified variables will have DummyCoding.

maxiter::Integer = 10000: Maximum number of iterations

tol::Real =1e-8: Tolerance

Details

Models with instruments variables are estimated using 2SLS. reg tests for weak instruments by computing the Kleibergen-Paap rk Wald F statistic, a generalization of the Cragg-Donald Wald F statistic for non i.i.d. errors. The statistic is similar to the one returned by the Stata command ivreg2.

Examples

using DataFrames, RDatasets, FixedEffectModels
df = dataset("plm", "Cigar")
df[:StateC] =  categorical(df[:State])
df[:YearC] =  categorical(df[:Year])
reg(df, @model(Sales ~ Price, fe = StateC + YearC))
reg(df, @model(Sales ~ NDI, fe = StateC + StateC&Year))
reg(df, @model(Sales ~ NDI, fe = StateC*Year))
reg(df, @model(Sales ~ (Price ~ Pimin)))
reg(df, @model(Sales ~ Price, weights = Pop))
reg(df, @model(Sales ~ NDI, subset = State .< 30))
reg(df, @model(Sales ~ NDI, vcov = robust))
reg(df, @model(Sales ~ NDI, vcov = cluster(StateC)))
reg(df, @model(Sales ~ NDI, vcov = cluster(StateC + YearC)))
reg(df, @model(Sales ~ YearC), contrasts = Dict(:YearC => DummyCoding(base = 80)))

FixedEffectModels.remove_singletons!

FixedEffectModels.renyientropy

FixedEffectModels.residuals

FixedEffectModels.response

FixedEffectModels.rle

FixedEffectModels.rmsd

FixedEffectModels.rss

FixedEffectModels.r²

FixedEffectModels.sample

FixedEffectModels.sample!

FixedEffectModels.samplepair

FixedEffectModels.sandwich

FixedEffectModels.scattermat

FixedEffectModels.score

FixedEffectModels.secondstage!

FixedEffectModels.sem

FixedEffectModels.setcontrasts!

FixedEffectModels.shat!

FixedEffectModels.skewness

FixedEffectModels.solve_coefficients!

FixedEffectModels.solve_residuals!

FixedEffectModels.span

FixedEffectModels.sqL2dist

FixedEffectModels.stderror

FixedEffectModels.summarystats

FixedEffectModels.tiedrank

FixedEffectModels.title

FixedEffectModels.top

FixedEffectModels.totalvar

FixedEffectModels.trim

FixedEffectModels.trim!

FixedEffectModels.trimvar

FixedEffectModels.variation

FixedEffectModels.vcov

FixedEffectModels.vcov!

FixedEffectModels.weights

FixedEffectModels.winsor

FixedEffectModels.winsor!

FixedEffectModels.wmean

FixedEffectModels.wmean!

FixedEffectModels.wmedian

FixedEffectModels.wquantile

FixedEffectModels.wsample

FixedEffectModels.wsample!

FixedEffectModels.wsum

FixedEffectModels.wsum!

FixedEffectModels.zscore

FixedEffectModels.zscore!