Docstrings

ReadWriteDlm2.ReadWriteDlm2 — Module.

ReadWriteDlm2

ReadWriteDlm2 functions readdlm2(), writedlm2(), readcsv2() and writecsv2() are similar to those of stdlib.DelimitedFiles, but with additional support for Date, DateTime, Time, Complex, Rational, Missing types and special decimal marks.

readcsv2(), writecsv2():

For "decimal dot" users the functions readcsv2() and writecsv2() have the respective defaults: Delimiter is ',' (fixed) and decimal='.'.

readdlm2(), writedlm2():

The basic idea of these functions is to support the "decimal comma countries". They use ';' as default delimiter and ',' as default decimal mark. "Decimal dot" users of these functions need to define decimal='.'

Detailed Documentation:

For more information about functionality and (keyword) arguments see ?help for readdlm2(), writedlm2(), readcsv2() and writecsv2().

ReadWriteDlm2.complexformat — Function.

Complexformat(a, decimal::AbstractChar, imsuffix::AbstractString)

Convert Complex number to String, optional change of decimal and/or imsuffix.

ReadWriteDlm2.dfregex — Function.

dfregex(df::AbstractString, locale::AbstractString="english")

Create a regex string r"^...$" for the given Date or DateTime formatstring df.

The regex groups are named according to the formatstring codes. locale is used to calculate min and max length of month and day names (for codes: UuEe).

ReadWriteDlm2.eval

ReadWriteDlm2.floatformat — Function.

floatformat(a, decimal::AbstractChar)

Convert Int or Float64 numbers to String and change decimal mark.

ReadWriteDlm2.include

ReadWriteDlm2.parseothers — Function.

parseothers(y::AbstractString, doparsetime::Bool, doparsecomplex::Bool, doparserational::Bool)

Parse string y for Time, Complex and Rational format and if match return the value. Otherwise return the input string y.

ReadWriteDlm2.readcsv2 — Function.

readcsv2(source, T::Type=Any; opts...)

Equivalent to readdlm2() with delimiter ',' and decimal='.'. Default Type Any activates parsing of Bool, Int, Float64, Complex, Rational, Missing, DateTime, Date and Time.

Code Example

julia> using ReadWriteDlm2

julia> B = Any[1 complex(1.5,2.7);1.0 1//3];

julia> writecsv2("test.csv", B)

julia> readcsv2("test.csv")
2×2 Array{Any,2}:
 1    1.5+2.7im
 1.0    1//3

ReadWriteDlm2.readdlm2 — Function.

readdlm2(source; options...)
readdlm2(source, T::Type; options...)
readdlm2(source, delim::AbstractChar; options...)
readdlm2(source, delim::AbstractChar, T::Type; options...)
readdlm2(source, delim::AbstractChar, eol::AbstractChar; options...)
readdlm2(source, delim::AbstractChar, T::Type, eol::AbstractChar; options...)

Read a matrix from source. The source can be a text file, stream or byte array. Each line (separated by eol, this is '\n' by default) gives one row. The columns are separated by ';', another delim can be defined.

Pre-processing of source with regex substitution changes the decimal marks from d,d to d.d. For default rs the keyword argument decimal=',' sets the decimal Char in the r-string of rs. When a special regex substitution tuple rs=(r.., s..) is defined, the argument decimal is not used. Pre-processing can be switched off with: rs=().

In addition to stdlib readdlm(), data is also parsed for Dates formats (ISO), the Time format "HH:MM[:SS[.s{1,9}]]" and for complex and rational numbers. To deactivate parsing dates/time set: dfs="", dtfs="". locale defines the language of day (E, e) and month (U, u) names.

The result will be a (heterogeneous) array of default type Any. Other (abstract) types for the array elements could be defined. If data is empty, a 0×0 Array{T,2} is returned. If dfheader=true instead of header=true is used, the returned tuple (data::Array{T,2}, header::Array{Symbol,1}) fits for DataFrames.

Additional Keyword Arguments

decimal=',': Decimal mark Char used by default rs, irrelevant if rs-tuple is not the default one

rs=(r"(\d),(\d)", s"\1.\2"): Regex (r,s)-tuple, the default change d,d to d.d if decimal=','

dtfs="yyyy-mm-ddTHH:MM:SS.s": Format string for DateTime parsing

dfs="yyyy-mm-dd": Format string for Date parsing

locale="english": Language for parsing dates names

dfheader=false: Return header in DataFrames format if true

missingstring="na": How missing values are represented

Find more information about readdlm() functionality and (keyword) arguments - which are also supported by readdlm2() - in help for readdlm().

Code Example

julia> using ReadWriteDlm2

julia> A = Any[1 1.2; "text" missing];

julia> writedlm2("test.csv", A)

julia> readdlm2("test.csv")
2×2 Array{Any,2}:
 1        1.2
  "text"   missing

ReadWriteDlm2.readdlm2auto

ReadWriteDlm2.timeformat — Function.

timeformat(a, decimal::AbstractChar)

Convert Time to String, optional with change of decimal mark for secounds.

ReadWriteDlm2.writecsv2 — Function.

writecsv2(f, A; opts...)

Equivalent to writedlm2() with fixed delimiter ',' and decimal='.'.

Code Example

julia> using ReadWriteDlm2

julia> B = Any[1 complex(1.5,2.7);1.0 1//3];

julia> writecsv2("test.csv", B)

julia> read("test.csv", String)
"1,1.5 + 2.7im\n1.0,1//3\n"

ReadWriteDlm2.writedlm2 — Function.

writedlm2(f, A; opts...)
writedlm2(f, A, delim; opts...)

Write A (a vector, matrix, or an iterable collection of iterable rows) as text to f(either a filename string or an IO stream). The columns will be separated by ';', another delim (Char or String) can be defined.

By default, a pre-processing of values takes place. Before writing as strings, decimal marks are changed from '.' to ','. With the keyword argument decimal= another decimal mark can be defined. To switch off this pre-processing set: decimal='.'.

In writedlm2() the output format for Date and DateTime data can be defined with format strings. Defaults are the ISO formats. Day (E, e) and month (U, u) names are written in the locale language. For writing Complex numbers the imaginary component suffix can be selected with the imsuffix= keyword argument.

Additional Keyword Arguments

decimal=',': Character for writing decimal marks

dtfs="yyyy-mm-ddTHH:MM:SS.s": DateTime write format

dfs="yyyy-mm-dd": Date write format

locale="english": Language for DateTime writing

imsuffix="im": Complex Imag suffix "im", "i" or "j"

missingstring="na": How missing values are written

Code Example

julia> using ReadWriteDlm2, Dates

julia> A = Any[1 1.2; "text" Date(2017)];

julia> writedlm2("test.csv", A)

julia> read("test.csv", String)
"1;1,2\ntext;2017-01-01\n"

ReadWriteDlm2.writedlm2auto