Docstrings

Watershed.AffinityMap

Watershed.DEFAULT_DUST_SIZE

Watershed.DEFAULT_HIGH

Watershed.DEFAULT_IS_THRESHOLD_RELATIVE

Watershed.DEFAULT_LOW

Watershed.DEFAULT_THRESHOLDS

Watershed.RegionGraph

Watershed.Segmentation

Watershed.Watershed

Watershed._baseseg

Watershed._percent2thd

Watershed._wsseg2d

Watershed.aff2sgm

Watershed.atomicseg

Watershed.divideplateaus! — Function.

DIVIDEPLATEAUS! - Divide plateaus in steepest ascent graph

 divideplateaus!(sag)

sag: steepest ascent graph (directed and unweighted). sag[x,y,z] contains 6-bit number encoding edges outgoing from (x,y,z)

Modify steepest ascent graph so as to

Divide non-maximal plateaus into paths that exit as quickly as possible

Break ties between multiple outgoing edges

Note this is an in-place modification of sag

Watershed.eval

Watershed.findbasins — Function.

FINDBASINS - find basins of attraction

 seg, counts, counts0 = findbasins(sag)

sag: steepest ascent graph (directed and unweighted). sag[x,y,z] contains 6-bit number encoding edges outgoing from (x,y,z)

seg: segmentation into basins. Each element of the 3D array contains a basin ID, a nonnegative integer ranging from 0 to the number of basins.

counts: number of voxels in each basin

counts0: number of background voxels

A value of 0 in seg indicates a background voxel, which has no edges at all in the steepest ascent graph. All such singletons are given the same ID of 0, although they are technically basins by themselves.

The algorithm starts from an unassigned voxel, and identifies all downstream voxels via breadth-first search (BFS). The search terminates in two possible ways:

downstream voxel that was previously assigned a basin ID =>

assign that ID to queued voxels.

no more downstream voxels => assign new basin ID to queued voxels.

Then the queue is emptied, and BFS begins anew at an unassigned voxel. The algorithm ends when all voxels are assigned.

The 7th bit (0x40) is used to indicate whether a voxel has been visited during BFS.

The MSB indicates whether a voxel has been assigned a basin ID. The MSB definition is given in the functions at the end of the file for UInt32 and UInt64 cases.

findbasins is applied to the steepest ascent graph after modification by divideplateaus! By this point all paths are unique, except in maximal plateaus.

what happens if findbasins is applied directly to the output of steepestascent? ties are resolved in an unsystematic way. seems to differ from Cousty's watershed cuts, which alternates btw DFS and BFS.

Watershed.findbasins!

Watershed.high_bit

Watershed.include

Watershed.low_bits

Watershed.mergeregions! — Function.

MERGEREGIONS - merge small regions by agglomerative clustering

new_rg = mergeregions(seg, rg, counts, thresholds, dust_size = 0)

seg - segmentation. IDs of foreground regions are 1:length(counts). ID=0 for background. This is modified in place by the clustering.

rg: region graph as list of edges, array of (weight,id1,id2) tuples. The edges should be presorted so that weights are in descending order. Region IDs should be consistent with those in seg, except no zeros.

new_rg: new region graph after clustering, same format as rg.

counts: sizes of regions in seg (modified in place)

thresholds: sequence of (sizeth,weightth) pairs to be used for merging

dust_size: after merging, tiny regions less than dust_size to be eliminated by changing them to background voxels

Agglomerative clustering proceeds by considering the edges of the region graph in sequence. If either region has size less than size_th, then merge the regions. When the weight of the edge in the region graph is less than or equal to weight_th, agglomeration proceeds to the next (size_th,weight_th) in thresholds or terminates if there is none.

to-do: update code to include self-edges in new_rg

Watershed.mergerg

Watershed.mergerg!

Watershed.mst — Function.

MST - compute maximal spanning tree from weighted graph

 regiontree = mst(rg,max_segid)

rg: region graph as list of edges, array of (weight,id1,id2) tuples. The edges should be presorted so that weights are in descending order.

max_segid: largest ID in region graph

regiontree: maximal spanning tree (MST) of region graph as list of edges, array of (weight,id1,id2) tuples. The vertices in each edge are ordered so that id2 is unique across edges. In other words, id1 and id2 correspond to parent and child in the tree. The code places the root of the tree at segid=1

The MST effectively represents the segmentPairsrogram for single-linkage clustering. Each edge (weight,id1,id2) in the MST represents an internal vertex of the dendrogram located at height = weight, i.e., a merging of two clusters with score weight. The MST contains a bit more information than the dendrogram, because id1 and id2 are the root IDs of the two clusters, i.e., the maximum weight edge between the two clusters is between their elements id1 and id2.

The code should work for general graphs. If edges of rg are presorted by ascending weight, the code will compute the minimal spanning tree rather than the maximal spanning tree.

Watershed.regiongraph — Function.

REGIONGRAPH - create region graph by finding maximum affinity between each pair of regions in segmentation

 rg = regiongraph(aff,seg,max_segid)

rg: region graph as list of edges, array of (weight,id1,id2) tuples. The edges are sorted so that weights are in descending order.

aff: affinity graph (undirected and weighted). 4D array of affinities, where last dimension is of size 3

seg: segmentation. Each element of the 3D array contains a segment ID, a nonnegative integer ranging from 0 to max_segid

max_segid: number of segments

The vertices of the region graph are regions in the segmentation. An edge of the region graph corresponds to a pair of regions in the segmentation that are connected by an edge in the affinity graph. The weight of an edge in the region graph is the maximum weight of the edges in the affinity graph connecting the two regions.

The region graph includes every edge between a region and itself. The weight of a self-edge is the maximum affinity within the region.

Background voxels (those with ID=0) are ignored.

Watershed.relative2absolute

Watershed.rg2segmentPairs — Function.

transform rg to dendrogram for omnification

Watershed.steepestascent — Function.

STEEPESTASCENT - Construct steepest ascent graph from affinity graph

 sag = steepestascent(aff, low, high)

sag: steepest ascent graph (directed and unweighted). sag[x,y,z] contains 6-bit number encoding edges outgoing from (x,y,z)

aff: affinity graph (undirected and weighted). 4D array of affinities, where last dimension is of size 3

low: edges with affinity <= low are removed

high: affinities >= high are considered infinity

Directed paths in the steepest ascent graph are steepest ascent paths in the affinity graph. Both graphs are for 3D lattice with 6-connectivity. The steepest ascent graph can contain vertices with multiple outgoing edges if there are ties in the affinity graph, i.e., if steepest ascent paths are nonunique.

We follow the convention that:

aff[x,y,z,1] is affinity of voxels at [x-1,y,z] and [x,y,z]

aff[x,y,z,2] is affinity of voxels at [x,y-1,z] and [x,y,z]

aff[x,y,z,3] is affinity of voxels at [x,y,z-1] and [x,y,z]

Watershed.watershed

Watershed.wsseg