Docstrings

StatisticalRethinking.+

StatisticalRethinking.-

StatisticalRethinking.@animate

StatisticalRethinking.@colorant_str

StatisticalRethinking.@csv2_str

StatisticalRethinking.@csv_str

StatisticalRethinking.@df

StatisticalRethinking.@gif

StatisticalRethinking.@layout

StatisticalRethinking.@recipe

StatisticalRethinking.@series

StatisticalRethinking.@shorthands

StatisticalRethinking.@tsv_str

StatisticalRethinking.@userplot

StatisticalRethinking.@wsv_str

StatisticalRethinking.ABGR

StatisticalRethinking.ADIN99

StatisticalRethinking.ADIN99d

StatisticalRethinking.ADIN99o

StatisticalRethinking.AGray

StatisticalRethinking.AGray32

StatisticalRethinking.AHSI

StatisticalRethinking.AHSL

StatisticalRethinking.AHSV

StatisticalRethinking.ALCHab

StatisticalRethinking.ALCHuv

StatisticalRethinking.ALMS

StatisticalRethinking.ALab

StatisticalRethinking.ALuv

StatisticalRethinking.ARGB

StatisticalRethinking.ARGB32

StatisticalRethinking.AXYZ

StatisticalRethinking.AYCbCr

StatisticalRethinking.AYIQ

StatisticalRethinking.AbstractBackend

StatisticalRethinking.AbstractCategoricalArray

StatisticalRethinking.AbstractCategoricalMatrix

StatisticalRethinking.AbstractCategoricalVector

StatisticalRethinking.AbstractDataFrame

StatisticalRethinking.AbstractGray

StatisticalRethinking.AbstractHistogram

StatisticalRethinking.AbstractLayout

StatisticalRethinking.AbstractMissingCategoricalArray

StatisticalRethinking.AbstractMissingCategoricalMatrix

StatisticalRethinking.AbstractMissingCategoricalVector

StatisticalRethinking.AbstractMixtureModel

StatisticalRethinking.AbstractMvNormal

StatisticalRethinking.AbstractPlot

StatisticalRethinking.AbstractRGB

StatisticalRethinking.AbstractSerializer

StatisticalRethinking.AbstractWeights

StatisticalRethinking.AlphaColor

StatisticalRethinking.AnalyticWeights

StatisticalRethinking.Animation

StatisticalRethinking.Arcsine

StatisticalRethinking.AxyY

StatisticalRethinking.BGR

StatisticalRethinking.BGRA

StatisticalRethinking.Bernoulli

StatisticalRethinking.Beta

StatisticalRethinking.BetaBinomial

StatisticalRethinking.BetaPrime

StatisticalRethinking.BezierCurve

StatisticalRethinking.Binomial

StatisticalRethinking.Biweight

StatisticalRethinking.CIE1931JV_CMF

StatisticalRethinking.CIE1931J_CMF

StatisticalRethinking.CIE1931_CMF

StatisticalRethinking.CIE1964_CMF

StatisticalRethinking.CSV

StatisticalRethinking.Categorical

StatisticalRethinking.CategoricalArray

StatisticalRethinking.CategoricalArrays

StatisticalRethinking.CategoricalMatrix

StatisticalRethinking.CategoricalPool

StatisticalRethinking.CategoricalString

StatisticalRethinking.CategoricalValue

StatisticalRethinking.CategoricalVector

StatisticalRethinking.Cauchy

StatisticalRethinking.Chains

StatisticalRethinking.Chi

StatisticalRethinking.Chisq

StatisticalRethinking.CoefTable

StatisticalRethinking.Color

StatisticalRethinking.Color3

StatisticalRethinking.ColorAlpha

StatisticalRethinking.ColorGradient

StatisticalRethinking.ColorTypes

StatisticalRethinking.Colorant

StatisticalRethinking.ColorantNormed

StatisticalRethinking.Colors

StatisticalRethinking.Continuous

StatisticalRethinking.ContinuousDistribution

StatisticalRethinking.ContinuousMatrixDistribution

StatisticalRethinking.ContinuousMultivariateDistribution

StatisticalRethinking.ContinuousUnivariateDistribution

StatisticalRethinking.ConvergenceException

StatisticalRethinking.Cosine

StatisticalRethinking.DE_2000

StatisticalRethinking.DE_94

StatisticalRethinking.DE_AB

StatisticalRethinking.DE_BFD

StatisticalRethinking.DE_CMC

StatisticalRethinking.DE_DIN99

StatisticalRethinking.DE_DIN99d

StatisticalRethinking.DE_DIN99o

StatisticalRethinking.DE_JPC79

StatisticalRethinking.DIN99

StatisticalRethinking.DIN99A

StatisticalRethinking.DIN99d

StatisticalRethinking.DIN99dA

StatisticalRethinking.DIN99o

StatisticalRethinking.DIN99oA

StatisticalRethinking.DataFrame

StatisticalRethinking.DataFrameRow

StatisticalRethinking.DataFrames

StatisticalRethinking.DelimitedFiles

StatisticalRethinking.DiagNormal

StatisticalRethinking.DiagNormalCanon

StatisticalRethinking.Dirichlet

StatisticalRethinking.DirichletMultinomial

StatisticalRethinking.Discrete

StatisticalRethinking.DiscreteDistribution

StatisticalRethinking.DiscreteMatrixDistribution

StatisticalRethinking.DiscreteMultivariateDistribution

StatisticalRethinking.DiscreteUniform

StatisticalRethinking.DiscreteUnivariateDistribution

StatisticalRethinking.Distribution

StatisticalRethinking.Distributions

StatisticalRethinking.DoubleExponential

StatisticalRethinking.ECDF

StatisticalRethinking.EdgeworthMean

StatisticalRethinking.EdgeworthSum

StatisticalRethinking.EdgeworthZ

StatisticalRethinking.EmpiricalUnivariateDistribution

StatisticalRethinking.Epanechnikov

StatisticalRethinking.Erlang

StatisticalRethinking.Estimator

StatisticalRethinking.Exponential

StatisticalRethinking.FDist

StatisticalRethinking.FisherNoncentralHypergeometric

StatisticalRethinking.Formatted

StatisticalRethinking.Fractional

StatisticalRethinking.Frechet

StatisticalRethinking.FrequencyWeights

StatisticalRethinking.FullNormal

StatisticalRethinking.FullNormalCanon

StatisticalRethinking.GR

StatisticalRethinking.Gamma

StatisticalRethinking.GeneralizedExtremeValue

StatisticalRethinking.GeneralizedPareto

StatisticalRethinking.Geometric

StatisticalRethinking.Gray

StatisticalRethinking.Gray24

StatisticalRethinking.GrayA

StatisticalRethinking.GroupApplied

StatisticalRethinking.GroupedDataFrame

StatisticalRethinking.Gumbel

StatisticalRethinking.HSB

StatisticalRethinking.HSI

StatisticalRethinking.HSIA

StatisticalRethinking.HSL

StatisticalRethinking.HSLA

StatisticalRethinking.HSV

StatisticalRethinking.HSVA

StatisticalRethinking.Histogram

StatisticalRethinking.Hypergeometric

StatisticalRethinking.InverseGamma

StatisticalRethinking.InverseGaussian

StatisticalRethinking.InverseWishart

StatisticalRethinking.IsoNormal

StatisticalRethinking.IsoNormalCanon

StatisticalRethinking.KSDist

StatisticalRethinking.KSOneSided

StatisticalRethinking.KW

StatisticalRethinking.Kolmogorov

StatisticalRethinking.L1dist

StatisticalRethinking.L2dist

StatisticalRethinking.LCHab

StatisticalRethinking.LCHabA

StatisticalRethinking.LCHuv

StatisticalRethinking.LCHuvA

StatisticalRethinking.LMS

StatisticalRethinking.LMSA

StatisticalRethinking.Lab

StatisticalRethinking.LabA

StatisticalRethinking.Laplace

StatisticalRethinking.LevelsException

StatisticalRethinking.Levy

StatisticalRethinking.Linfdist

StatisticalRethinking.LocationScale

StatisticalRethinking.LogNormal

StatisticalRethinking.Logistic

StatisticalRethinking.Luv

StatisticalRethinking.LuvA

StatisticalRethinking.MCMCChain

StatisticalRethinking.MLEstimator

StatisticalRethinking.MSC

StatisticalRethinking.MatrixDistribution

StatisticalRethinking.Matrixvariate

StatisticalRethinking.Missing

StatisticalRethinking.MissingCategoricalArray

StatisticalRethinking.MissingCategoricalMatrix

StatisticalRethinking.MissingCategoricalVector

StatisticalRethinking.MissingException

StatisticalRethinking.Missings

StatisticalRethinking.MixtureModel

StatisticalRethinking.Multinomial

StatisticalRethinking.Multivariate

StatisticalRethinking.MultivariateDistribution

StatisticalRethinking.MultivariateMixture

StatisticalRethinking.MultivariateNormal

StatisticalRethinking.MvLogNormal

StatisticalRethinking.MvNormal

StatisticalRethinking.MvNormalCanon

StatisticalRethinking.MvNormalKnownCov

StatisticalRethinking.MvTDist

StatisticalRethinking.NegativeBinomial

StatisticalRethinking.NonMatrixDistribution

StatisticalRethinking.NoncentralBeta

StatisticalRethinking.NoncentralChisq

StatisticalRethinking.NoncentralF

StatisticalRethinking.NoncentralHypergeometric

StatisticalRethinking.NoncentralT

StatisticalRethinking.Normal

StatisticalRethinking.NormalCanon

StatisticalRethinking.NormalInverseGaussian

StatisticalRethinking.OHLC

StatisticalRethinking.OrderedLevelsException

StatisticalRethinking.Pareto

StatisticalRethinking.PlotThemes

StatisticalRethinking.PlotUtils

StatisticalRethinking.Plots

StatisticalRethinking.Poisson

StatisticalRethinking.PoissonBinomial

StatisticalRethinking.ProbabilityWeights

StatisticalRethinking.QQPair

StatisticalRethinking.RDatasets

StatisticalRethinking.RGB

StatisticalRethinking.RGB1

StatisticalRethinking.RGB24

StatisticalRethinking.RGB4

StatisticalRethinking.RGBA

StatisticalRethinking.Rayleigh

StatisticalRethinking.RealInterval

StatisticalRethinking.RecipeData

StatisticalRethinking.RecipesBase

StatisticalRethinking.RegressionModel

StatisticalRethinking.Sampleable

StatisticalRethinking.ScriptEntry — Type.

ScriptEntry

Define processing requirements for chapter scripts

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

StatisticalRethinking.Segments

StatisticalRethinking.Semicircle

StatisticalRethinking.Serialization

StatisticalRethinking.Serializer

StatisticalRethinking.Shape

StatisticalRethinking.Skellam

StatisticalRethinking.StatisticalModel

StatisticalRethinking.StatisticalRethinking

StatisticalRethinking.StatsBase

StatisticalRethinking.StatsFuns

StatisticalRethinking.StatsPlots

StatisticalRethinking.SubDataFrame

StatisticalRethinking.SufficientStats

StatisticalRethinking.Surface

StatisticalRethinking.SymTriangularDist

StatisticalRethinking.TDist

StatisticalRethinking.Transparent3

StatisticalRethinking.TransparentColor

StatisticalRethinking.TransparentGray

StatisticalRethinking.TransparentRGB

StatisticalRethinking.TriangularDist

StatisticalRethinking.Triweight

StatisticalRethinking.Truncated

StatisticalRethinking.TruncatedNormal

StatisticalRethinking.Uniform

StatisticalRethinking.Univariate

StatisticalRethinking.UnivariateDistribution

StatisticalRethinking.UnivariateGMM

StatisticalRethinking.UnivariateMixture

StatisticalRethinking.ValueSupport

StatisticalRethinking.VariateForm

StatisticalRethinking.VonMises

StatisticalRethinking.VonMisesFisher

StatisticalRethinking.WalleniusNoncentralHypergeometric

StatisticalRethinking.Weibull

StatisticalRethinking.Weights

StatisticalRethinking.Wishart

StatisticalRethinking.XYZ

StatisticalRethinking.XYZA

StatisticalRethinking.YCbCr

StatisticalRethinking.YCbCrA

StatisticalRethinking.YIQ

StatisticalRethinking.YIQA

StatisticalRethinking.ZeroMeanDiagNormal

StatisticalRethinking.ZeroMeanDiagNormalCanon

StatisticalRethinking.ZeroMeanFullNormal

StatisticalRethinking.ZeroMeanFullNormalCanon

StatisticalRethinking.ZeroMeanIsoNormal

StatisticalRethinking.ZeroMeanIsoNormalCanon

StatisticalRethinking.adapted_grid

StatisticalRethinking.add_backend

StatisticalRethinking.add_theme

StatisticalRethinking.addcounts!

StatisticalRethinking.adjr2

StatisticalRethinking.adjr²

StatisticalRethinking.aggregate

StatisticalRethinking.aic

StatisticalRethinking.aicc

StatisticalRethinking.aliases

StatisticalRethinking.allowmissing

StatisticalRethinking.allowmissing!

StatisticalRethinking.alpha

StatisticalRethinking.alphacolor

StatisticalRethinking.andrewsplot

StatisticalRethinking.andrewsplot!

StatisticalRethinking.animate

StatisticalRethinking.annotate!

StatisticalRethinking.arrow

StatisticalRethinking.attr!

StatisticalRethinking.autcorplot

StatisticalRethinking.autocor

StatisticalRethinking.autocor!

StatisticalRethinking.autocorplot

StatisticalRethinking.autocorplot!

StatisticalRethinking.autocov

StatisticalRethinking.autocov!

StatisticalRethinking.aweights

StatisticalRethinking.backend

StatisticalRethinking.backend_name

StatisticalRethinking.backend_object

StatisticalRethinking.backends

StatisticalRethinking.bar

StatisticalRethinking.bar!

StatisticalRethinking.barh

StatisticalRethinking.barh!

StatisticalRethinking.barhist

StatisticalRethinking.barhist!

StatisticalRethinking.base_color_type

StatisticalRethinking.base_colorant_type

StatisticalRethinking.bbox

StatisticalRethinking.betaccdf

StatisticalRethinking.betacdf

StatisticalRethinking.betainvccdf

StatisticalRethinking.betainvcdf

StatisticalRethinking.betainvlogccdf

StatisticalRethinking.betainvlogcdf

StatisticalRethinking.betalogccdf

StatisticalRethinking.betalogcdf

StatisticalRethinking.betalogpdf

StatisticalRethinking.betapdf

StatisticalRethinking.bic

StatisticalRethinking.binaryentropy

StatisticalRethinking.binomccdf

StatisticalRethinking.binomcdf

StatisticalRethinking.binominvccdf

StatisticalRethinking.binominvcdf

StatisticalRethinking.binominvlogccdf

StatisticalRethinking.binominvlogcdf

StatisticalRethinking.binomlogccdf

StatisticalRethinking.binomlogcdf

StatisticalRethinking.binomlogpdf

StatisticalRethinking.binompdf

StatisticalRethinking.blue

StatisticalRethinking.boxplot

StatisticalRethinking.boxplot!

StatisticalRethinking.brush

StatisticalRethinking.by

StatisticalRethinking.canonform

StatisticalRethinking.categorical

StatisticalRethinking.categorical!

StatisticalRethinking.ccdf

StatisticalRethinking.ccolor

StatisticalRethinking.cdensity

StatisticalRethinking.cdensity!

StatisticalRethinking.cdf

StatisticalRethinking.center

StatisticalRethinking.cf

StatisticalRethinking.cgf

StatisticalRethinking.cgrad

StatisticalRethinking.cgradients

StatisticalRethinking.chisqccdf

StatisticalRethinking.chisqcdf

StatisticalRethinking.chisqinvccdf

StatisticalRethinking.chisqinvcdf

StatisticalRethinking.chisqinvlogccdf

StatisticalRethinking.chisqinvlogcdf

StatisticalRethinking.chisqlogccdf

StatisticalRethinking.chisqlogcdf

StatisticalRethinking.chisqlogpdf

StatisticalRethinking.chisqpdf

StatisticalRethinking.cie_color_match

StatisticalRethinking.circvar

StatisticalRethinking.clibraries

StatisticalRethinking.clibrary

StatisticalRethinking.closeall

StatisticalRethinking.coalesce

StatisticalRethinking.coef

StatisticalRethinking.coefnames

StatisticalRethinking.coeftable

StatisticalRethinking.color

StatisticalRethinking.color_type

StatisticalRethinking.coloralpha

StatisticalRethinking.colordiff

StatisticalRethinking.colormap

StatisticalRethinking.colormatch

StatisticalRethinking.colwise

StatisticalRethinking.combine

StatisticalRethinking.comp1

StatisticalRethinking.comp2

StatisticalRethinking.comp3

StatisticalRethinking.compact

StatisticalRethinking.competerank

StatisticalRethinking.complete_cases

StatisticalRethinking.complete_cases!

StatisticalRethinking.completecases

StatisticalRethinking.component

StatisticalRethinking.components

StatisticalRethinking.componentwise_logpdf

StatisticalRethinking.componentwise_pdf

StatisticalRethinking.compress

StatisticalRethinking.concentration

StatisticalRethinking.confint

StatisticalRethinking.contour

StatisticalRethinking.contour!

StatisticalRethinking.contour3d

StatisticalRethinking.contour3d!

StatisticalRethinking.contourf

StatisticalRethinking.contourf!

StatisticalRethinking.convert

StatisticalRethinking.coords

StatisticalRethinking.cor

StatisticalRethinking.cor2cov

StatisticalRethinking.corkendall

StatisticalRethinking.corner

StatisticalRethinking.corner!

StatisticalRethinking.cornerplot

StatisticalRethinking.cornerplot!

StatisticalRethinking.corrplot

StatisticalRethinking.corrplot!

StatisticalRethinking.corspearman

StatisticalRethinking.counteq

StatisticalRethinking.countmap

StatisticalRethinking.countne

StatisticalRethinking.counts

StatisticalRethinking.cov

StatisticalRethinking.cov2cor

StatisticalRethinking.cquantile

StatisticalRethinking.crosscor

StatisticalRethinking.crosscor!

StatisticalRethinking.crosscov

StatisticalRethinking.crosscov!

StatisticalRethinking.crossentropy

StatisticalRethinking.cumulant

StatisticalRethinking.current

StatisticalRethinking.curve_points

StatisticalRethinking.curves

StatisticalRethinking.curves!

StatisticalRethinking.cut

StatisticalRethinking.cvec

StatisticalRethinking.dataset

StatisticalRethinking.dataviewer

StatisticalRethinking.decompress

StatisticalRethinking.default

StatisticalRethinking.default_cgrad

StatisticalRethinking.delete!

StatisticalRethinking.deletecols!

StatisticalRethinking.deleterows!

StatisticalRethinking.denserank

StatisticalRethinking.density

StatisticalRethinking.density!

StatisticalRethinking.densityplot

StatisticalRethinking.describe

StatisticalRethinking.deserialize

StatisticalRethinking.deuteranopic

StatisticalRethinking.deviance

StatisticalRethinking.dim

StatisticalRethinking.disallowmissing

StatisticalRethinking.disallowmissing!

StatisticalRethinking.discretediag

StatisticalRethinking.distinguishable_colors

StatisticalRethinking.diverging_palette

StatisticalRethinking.dof

StatisticalRethinking.dof_residual

StatisticalRethinking.droplevels!

StatisticalRethinking.dropmissing

StatisticalRethinking.dropmissing!

StatisticalRethinking.ea_histogram

StatisticalRethinking.ea_histogram!

StatisticalRethinking.ecdf

StatisticalRethinking.eltypes

StatisticalRethinking.entropy

StatisticalRethinking.estimate

StatisticalRethinking.eval

StatisticalRethinking.expected_logdet

StatisticalRethinking.failprob

StatisticalRethinking.fdistccdf

StatisticalRethinking.fdistcdf

StatisticalRethinking.fdistinvccdf

StatisticalRethinking.fdistinvcdf

StatisticalRethinking.fdistinvlogccdf

StatisticalRethinking.fdistinvlogcdf

StatisticalRethinking.fdistlogccdf

StatisticalRethinking.fdistlogcdf

StatisticalRethinking.fdistlogpdf

StatisticalRethinking.fdistpdf

StatisticalRethinking.findat

StatisticalRethinking.fit

StatisticalRethinking.fit!

StatisticalRethinking.fit_map

StatisticalRethinking.fit_map!

StatisticalRethinking.fit_mle

StatisticalRethinking.fit_mle!

StatisticalRethinking.fitted

StatisticalRethinking.font

StatisticalRethinking.fourinvπ

StatisticalRethinking.fourπ

StatisticalRethinking.frame

StatisticalRethinking.freecumulant

StatisticalRethinking.fweights

StatisticalRethinking.gammaccdf

StatisticalRethinking.gammacdf

StatisticalRethinking.gammainvccdf

StatisticalRethinking.gammainvcdf

StatisticalRethinking.gammainvlogccdf

StatisticalRethinking.gammainvlogcdf

StatisticalRethinking.gammalogccdf

StatisticalRethinking.gammalogcdf

StatisticalRethinking.gammalogpdf

StatisticalRethinking.gammapdf

StatisticalRethinking.gamutmax

StatisticalRethinking.gamutmin

StatisticalRethinking.gelmandiag

StatisticalRethinking.generate — Function.

generate

Utility function to generate all notebooks and chapters from scripts in the scripts directory.

Method

generate(sd = script_dict)

Required arguments

None, all notebooks/.. and chapters/.. files are regenerated.

generate

Generate notebooks and scripts in a single chapter.

Method

generate(chapter::AbstractString)

Required arguments

Generate notebooks and scripts in a single chapter, e.g. generate("04")

generate

Generate a single notebook and script

Method

generate(chapter::AbstractString, file::AbstractString)

Required arguments

Generate notebook and script file in chapter, e.g. generate("04", "m4.1d.jl") or generate("04/m4.1d.jl")

StatisticalRethinking.genmean

StatisticalRethinking.genvar

StatisticalRethinking.geomean

StatisticalRethinking.get_color_palette

StatisticalRethinking.getindex

StatisticalRethinking.gewekediag

StatisticalRethinking.gif

StatisticalRethinking.gkldiv

StatisticalRethinking.gr

StatisticalRethinking.gradlogpdf

StatisticalRethinking.gray

StatisticalRethinking.green

StatisticalRethinking.grid

StatisticalRethinking.groupby

StatisticalRethinking.groupedbar

StatisticalRethinking.groupedbar!

StatisticalRethinking.groupindices

StatisticalRethinking.groupvars

StatisticalRethinking.gui

StatisticalRethinking.halfπ

StatisticalRethinking.harmmean

StatisticalRethinking.hasfinitesupport

StatisticalRethinking.hdf5

StatisticalRethinking.head

StatisticalRethinking.heatmap

StatisticalRethinking.heatmap!

StatisticalRethinking.heideldiag

StatisticalRethinking.hex

StatisticalRethinking.hexbin

StatisticalRethinking.hexbin!

StatisticalRethinking.hist

StatisticalRethinking.histogram

StatisticalRethinking.histogram!

StatisticalRethinking.histogram2d

StatisticalRethinking.histogram2d!

StatisticalRethinking.histogramplot

StatisticalRethinking.hline

StatisticalRethinking.hline!

StatisticalRethinking.hspan

StatisticalRethinking.hspan!

StatisticalRethinking.hyperccdf

StatisticalRethinking.hypercdf

StatisticalRethinking.hyperinvccdf

StatisticalRethinking.hyperinvcdf

StatisticalRethinking.hyperinvlogccdf

StatisticalRethinking.hyperinvlogcdf

StatisticalRethinking.hyperlogccdf

StatisticalRethinking.hyperlogcdf

StatisticalRethinking.hyperlogpdf

StatisticalRethinking.hyperpdf

StatisticalRethinking.include

StatisticalRethinking.indexmap

StatisticalRethinking.indicatormat

StatisticalRethinking.informationmatrix

StatisticalRethinking.inline

StatisticalRethinking.insert!

StatisticalRethinking.insertcols!

StatisticalRethinking.inspectdr

StatisticalRethinking.insupport

StatisticalRethinking.inv2π

StatisticalRethinking.inv4π

StatisticalRethinking.invcov

StatisticalRethinking.inverse_rle

StatisticalRethinking.invisible

StatisticalRethinking.invlogccdf

StatisticalRethinking.invlogcdf

StatisticalRethinking.invscale

StatisticalRethinking.invsoftplus

StatisticalRethinking.invsqrt2

StatisticalRethinking.invsqrt2π

StatisticalRethinking.invπ

StatisticalRethinking.iqr

StatisticalRethinking.isbounded

StatisticalRethinking.isdark

StatisticalRethinking.isfitted

StatisticalRethinking.isleptokurtic

StatisticalRethinking.islinear

StatisticalRethinking.islowerbounded

StatisticalRethinking.ismesokurtic

StatisticalRethinking.ismissing

StatisticalRethinking.isordered

StatisticalRethinking.isplatykurtic

StatisticalRethinking.isprobvec

StatisticalRethinking.isupperbounded

StatisticalRethinking.iter_segments

StatisticalRethinking.kde

StatisticalRethinking.keys

StatisticalRethinking.kldivergence

StatisticalRethinking.kurtosis

StatisticalRethinking.length

StatisticalRethinking.levels

StatisticalRethinking.levels!

StatisticalRethinking.levelsmap

StatisticalRethinking.leverage

StatisticalRethinking.link — Function.

link

Compute the link function

Method

link(xrange, chain, vars, xbar) 

Required arguments

* `xrange::Turing model`  : Range over which link values are computed
* `chain::Float64`             : Chain samples used
* `vars::Float64`               : Variables in chain used
* `xbar::Float64`               : Mean value of observed predictor

Return values

* `result`                       : Vector of link values

StatisticalRethinking.location

StatisticalRethinking.location!

StatisticalRethinking.log1mexp

StatisticalRethinking.log1pexp

StatisticalRethinking.log1pmx

StatisticalRethinking.log1psq

StatisticalRethinking.log2mexp

StatisticalRethinking.log2π

StatisticalRethinking.log4π

StatisticalRethinking.logaddexp

StatisticalRethinking.logccdf

StatisticalRethinking.logcdf

StatisticalRethinking.logdetcov

StatisticalRethinking.logexpm1

StatisticalRethinking.loghalf

StatisticalRethinking.logistic

StatisticalRethinking.logit

StatisticalRethinking.loglikelihood

StatisticalRethinking.logmvgamma

StatisticalRethinking.logmxp1

StatisticalRethinking.logpdf

StatisticalRethinking.logpdf!

StatisticalRethinking.logsumexp

StatisticalRethinking.logtwo

StatisticalRethinking.logπ

StatisticalRethinking.lstirling_asym

StatisticalRethinking.mad

StatisticalRethinking.mad!

StatisticalRethinking.map

StatisticalRethinking.mapc

StatisticalRethinking.mapcols

StatisticalRethinking.mapreducec

StatisticalRethinking.marginalhist

StatisticalRethinking.marginalhist!

StatisticalRethinking.maxad

StatisticalRethinking.mean

StatisticalRethinking.mean_and_cov

StatisticalRethinking.mean_and_std

StatisticalRethinking.mean_and_var

StatisticalRethinking.meanad

StatisticalRethinking.meandir

StatisticalRethinking.meanform

StatisticalRethinking.meanlogx

StatisticalRethinking.meanplot

StatisticalRethinking.meanplot!

StatisticalRethinking.meanresponse

StatisticalRethinking.median

StatisticalRethinking.melt

StatisticalRethinking.meltdf

StatisticalRethinking.merge!

StatisticalRethinking.mgf

StatisticalRethinking.middle

StatisticalRethinking.midpoints

StatisticalRethinking.missing

StatisticalRethinking.missings

StatisticalRethinking.mixeddensity

StatisticalRethinking.mixeddensity!

StatisticalRethinking.mixeddensityplot

StatisticalRethinking.mode

StatisticalRethinking.model_response

StatisticalRethinking.modelmatrix

StatisticalRethinking.modes

StatisticalRethinking.moment

StatisticalRethinking.mov

StatisticalRethinking.mp4

StatisticalRethinking.msd

StatisticalRethinking.mss

StatisticalRethinking.names!

StatisticalRethinking.nbetaccdf

StatisticalRethinking.nbetacdf

StatisticalRethinking.nbetainvccdf

StatisticalRethinking.nbetainvcdf

StatisticalRethinking.nbetainvlogccdf

StatisticalRethinking.nbetainvlogcdf

StatisticalRethinking.nbetalogccdf

StatisticalRethinking.nbetalogcdf

StatisticalRethinking.nbetalogpdf

StatisticalRethinking.nbetapdf

StatisticalRethinking.nbinomccdf

StatisticalRethinking.nbinomcdf

StatisticalRethinking.nbinominvccdf

StatisticalRethinking.nbinominvcdf

StatisticalRethinking.nbinominvlogccdf

StatisticalRethinking.nbinominvlogcdf

StatisticalRethinking.nbinomlogccdf

StatisticalRethinking.nbinomlogcdf

StatisticalRethinking.nbinomlogpdf

StatisticalRethinking.nbinompdf

StatisticalRethinking.ncategories

StatisticalRethinking.nchisqccdf

StatisticalRethinking.nchisqcdf

StatisticalRethinking.nchisqinvccdf

StatisticalRethinking.nchisqinvcdf

StatisticalRethinking.nchisqinvlogccdf

StatisticalRethinking.nchisqinvlogcdf

StatisticalRethinking.nchisqlogccdf

StatisticalRethinking.nchisqlogcdf

StatisticalRethinking.nchisqlogpdf

StatisticalRethinking.nchisqpdf

StatisticalRethinking.ncol

StatisticalRethinking.ncomponents

StatisticalRethinking.nfdistccdf

StatisticalRethinking.nfdistcdf

StatisticalRethinking.nfdistinvccdf

StatisticalRethinking.nfdistinvcdf

StatisticalRethinking.nfdistinvlogccdf

StatisticalRethinking.nfdistinvlogcdf

StatisticalRethinking.nfdistlogccdf

StatisticalRethinking.nfdistlogcdf

StatisticalRethinking.nfdistlogpdf

StatisticalRethinking.nfdistpdf

StatisticalRethinking.nobs

StatisticalRethinking.nonunique

StatisticalRethinking.norepeats

StatisticalRethinking.normccdf

StatisticalRethinking.normcdf

StatisticalRethinking.norminvccdf

StatisticalRethinking.norminvcdf

StatisticalRethinking.norminvlogccdf

StatisticalRethinking.norminvlogcdf

StatisticalRethinking.normlogccdf

StatisticalRethinking.normlogcdf

StatisticalRethinking.normlogpdf

StatisticalRethinking.normpdf

StatisticalRethinking.nquantile

StatisticalRethinking.nrow

StatisticalRethinking.nsamples

StatisticalRethinking.ntdistccdf

StatisticalRethinking.ntdistcdf

StatisticalRethinking.ntdistinvccdf

StatisticalRethinking.ntdistinvcdf

StatisticalRethinking.ntdistinvlogccdf

StatisticalRethinking.ntdistinvlogcdf

StatisticalRethinking.ntdistlogccdf

StatisticalRethinking.ntdistlogcdf

StatisticalRethinking.ntdistlogpdf

StatisticalRethinking.ntdistpdf

StatisticalRethinking.ntrials

StatisticalRethinking.nullable!

StatisticalRethinking.nulldeviance

StatisticalRethinking.nullloglikelihood

StatisticalRethinking.ohlc

StatisticalRethinking.ohlc!

StatisticalRethinking.optimize_datetime_ticks

StatisticalRethinking.optimize_ticks

StatisticalRethinking.order

StatisticalRethinking.ordered

StatisticalRethinking.ordered!

StatisticalRethinking.ordinalrank

StatisticalRethinking.pacf

StatisticalRethinking.pacf!

StatisticalRethinking.palette

StatisticalRethinking.params

StatisticalRethinking.params!

StatisticalRethinking.partialcor

StatisticalRethinking.partype

StatisticalRethinking.passmissing

StatisticalRethinking.path3d

StatisticalRethinking.path3d!

StatisticalRethinking.pdf

StatisticalRethinking.percentile

StatisticalRethinking.permutecols!

StatisticalRethinking.pgfplots

StatisticalRethinking.pie

StatisticalRethinking.pie!

StatisticalRethinking.plot

StatisticalRethinking.plot!

StatisticalRethinking.plot3d

StatisticalRethinking.plot3d!

StatisticalRethinking.plot_color

StatisticalRethinking.plotarea

StatisticalRethinking.plotattr

StatisticalRethinking.plotly

StatisticalRethinking.plotlyjs

StatisticalRethinking.plots_heatmap

StatisticalRethinking.plots_heatmap!

StatisticalRethinking.png

StatisticalRethinking.poisccdf

StatisticalRethinking.poiscdf

StatisticalRethinking.poisinvccdf

StatisticalRethinking.poisinvcdf

StatisticalRethinking.poisinvlogccdf

StatisticalRethinking.poisinvlogcdf

StatisticalRethinking.poislogccdf

StatisticalRethinking.poislogcdf

StatisticalRethinking.poislogpdf

StatisticalRethinking.poispdf

StatisticalRethinking.pool

StatisticalRethinking.pool!

StatisticalRethinking.portfoliocomposition

StatisticalRethinking.portfoliocomposition!

StatisticalRethinking.predict

StatisticalRethinking.predict!

StatisticalRethinking.probs

StatisticalRethinking.probval

StatisticalRethinking.proportionmap

StatisticalRethinking.proportions

StatisticalRethinking.protanopic

StatisticalRethinking.psnr

StatisticalRethinking.pweights

StatisticalRethinking.pyplot

StatisticalRethinking.qqbuild

StatisticalRethinking.qqnorm

StatisticalRethinking.qqnorm!

StatisticalRethinking.qqplot

StatisticalRethinking.qqplot!

StatisticalRethinking.quantile

StatisticalRethinking.quartπ

StatisticalRethinking.quiver

StatisticalRethinking.quiver!

StatisticalRethinking.r2

StatisticalRethinking.rafterydiag

StatisticalRethinking.rate

StatisticalRethinking.readdlm

StatisticalRethinking.readtable

StatisticalRethinking.recode

StatisticalRethinking.recode!

StatisticalRethinking.red

StatisticalRethinking.reducec

StatisticalRethinking.rel_path — Function.

rel_path

Relative path using the StatisticalRethinking src/ directory. Copied from DynamicHMCExamples.jl

Example to get access to the data subdirectory

rel_path("..", "data")

StatisticalRethinking.rename

StatisticalRethinking.rename!

StatisticalRethinking.renyientropy

StatisticalRethinking.residuals

StatisticalRethinking.response

StatisticalRethinking.rgb_string

StatisticalRethinking.rgba_string

StatisticalRethinking.rle

StatisticalRethinking.rmsd

StatisticalRethinking.rotate

StatisticalRethinking.rotate!

StatisticalRethinking.rss

StatisticalRethinking.r²

StatisticalRethinking.sample

StatisticalRethinking.sample!

StatisticalRethinking.samplepair

StatisticalRethinking.sampler

StatisticalRethinking.savefig

StatisticalRethinking.scale

StatisticalRethinking.scale!

StatisticalRethinking.scatter

StatisticalRethinking.scatter!

StatisticalRethinking.scatter3d

StatisticalRethinking.scatter3d!

StatisticalRethinking.scatterhist

StatisticalRethinking.scatterhist!

StatisticalRethinking.scattermat

StatisticalRethinking.score

StatisticalRethinking.script_dict

StatisticalRethinking.scriptentry — Function.

scriptentry

Constructor for ScriptEntry objects.

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

StatisticalRethinking.sem

StatisticalRethinking.sequential_palette

StatisticalRethinking.serialize

StatisticalRethinking.set_theme

StatisticalRethinking.setindex!

StatisticalRethinking.shape

StatisticalRethinking.shape_coords

StatisticalRethinking.show

StatisticalRethinking.showall

StatisticalRethinking.showcols

StatisticalRethinking.showgradient

StatisticalRethinking.showgradient!

StatisticalRethinking.showlibrary

StatisticalRethinking.showlibrary!

StatisticalRethinking.showtheme

StatisticalRethinking.showtheme!

StatisticalRethinking.skewness

StatisticalRethinking.softmax

StatisticalRethinking.softmax!

StatisticalRethinking.softplus

StatisticalRethinking.span

StatisticalRethinking.spy

StatisticalRethinking.spy!

StatisticalRethinking.sqL2dist

StatisticalRethinking.sqmahal

StatisticalRethinking.sqmahal!

StatisticalRethinking.sqrt2

StatisticalRethinking.sqrt2π

StatisticalRethinking.sqrt3

StatisticalRethinking.sqrt4π

StatisticalRethinking.sqrthalfπ

StatisticalRethinking.sqrtπ

StatisticalRethinking.src_path

StatisticalRethinking.srdistccdf

StatisticalRethinking.srdistcdf

StatisticalRethinking.srdistinvccdf

StatisticalRethinking.srdistinvcdf

StatisticalRethinking.srdistinvlogccdf

StatisticalRethinking.srdistinvlogcdf

StatisticalRethinking.srdistlogccdf

StatisticalRethinking.srdistlogcdf

StatisticalRethinking.stack

StatisticalRethinking.stackdf

StatisticalRethinking.std

StatisticalRethinking.stderror

StatisticalRethinking.stdlogx

StatisticalRethinking.stephist

StatisticalRethinking.stephist!

StatisticalRethinking.sticks

StatisticalRethinking.sticks!

StatisticalRethinking.stroke

StatisticalRethinking.sub

StatisticalRethinking.succprob

StatisticalRethinking.suffstats

StatisticalRethinking.summarystats

StatisticalRethinking.support

StatisticalRethinking.surface

StatisticalRethinking.surface!

StatisticalRethinking.tail

StatisticalRethinking.tdistccdf

StatisticalRethinking.tdistcdf

StatisticalRethinking.tdistinvccdf

StatisticalRethinking.tdistinvcdf

StatisticalRethinking.tdistinvlogccdf

StatisticalRethinking.tdistinvlogcdf

StatisticalRethinking.tdistlogccdf

StatisticalRethinking.tdistlogcdf

StatisticalRethinking.tdistlogpdf

StatisticalRethinking.tdistpdf

StatisticalRethinking.test_distr

StatisticalRethinking.test_examples

StatisticalRethinking.test_samples

StatisticalRethinking.testhist

StatisticalRethinking.testhist!

StatisticalRethinking.text

StatisticalRethinking.theme

StatisticalRethinking.tiedrank

StatisticalRethinking.title!

StatisticalRethinking.totalvar

StatisticalRethinking.traceplot

StatisticalRethinking.traceplot!

StatisticalRethinking.transform

StatisticalRethinking.translate

StatisticalRethinking.translate!

StatisticalRethinking.trim

StatisticalRethinking.trim!

StatisticalRethinking.trimvar

StatisticalRethinking.tritanopic

StatisticalRethinking.twinx

StatisticalRethinking.twoinvπ

StatisticalRethinking.twoπ

StatisticalRethinking.uncompact

StatisticalRethinking.unicodeplots

StatisticalRethinking.unique!

StatisticalRethinking.unstack

StatisticalRethinking.values

StatisticalRethinking.var

StatisticalRethinking.variation

StatisticalRethinking.varlogx

StatisticalRethinking.vcat

StatisticalRethinking.vcov

StatisticalRethinking.violin

StatisticalRethinking.violin!

StatisticalRethinking.vline

StatisticalRethinking.vline!

StatisticalRethinking.vspan

StatisticalRethinking.vspan!

StatisticalRethinking.weighted_color_mean

StatisticalRethinking.weights

StatisticalRethinking.whitebalance

StatisticalRethinking.winsor

StatisticalRethinking.winsor!

StatisticalRethinking.wireframe

StatisticalRethinking.wireframe!

StatisticalRethinking.with

StatisticalRethinking.wmean

StatisticalRethinking.wmean!

StatisticalRethinking.wmedian

StatisticalRethinking.wquantile

StatisticalRethinking.wrap

StatisticalRethinking.writedlm

StatisticalRethinking.writetable

StatisticalRethinking.wsample

StatisticalRethinking.wsample!

StatisticalRethinking.wsum

StatisticalRethinking.wsum!

StatisticalRethinking.xaxis!

StatisticalRethinking.xflip!

StatisticalRethinking.xgrid!

StatisticalRethinking.xlabel!

StatisticalRethinking.xlims

StatisticalRethinking.xlims!

StatisticalRethinking.xlogx

StatisticalRethinking.xlogy

StatisticalRethinking.xticks!

StatisticalRethinking.xyY

StatisticalRethinking.xyYA

StatisticalRethinking.yaxis!

StatisticalRethinking.yflip!

StatisticalRethinking.ygrid!

StatisticalRethinking.ylabel!

StatisticalRethinking.ylims

StatisticalRethinking.ylims!

StatisticalRethinking.yticks!

StatisticalRethinking.zlims

StatisticalRethinking.zlims!

StatisticalRethinking.zscore

StatisticalRethinking.zscore!