Docstrings

Taro.CELL_TYPE_BLANK

Taro.CELL_TYPE_BOOLEAN

Taro.CELL_TYPE_ERROR

Taro.CELL_TYPE_FORMULA

Taro.CELL_TYPE_NUMERIC

Taro.CELL_TYPE_STRING

Taro.Cell — Type.

A cell within an excel sheet. Most operations to get or set values occur on a cell. Wrapper for Java class org.apache.poi.ss.usermodel.Cell

Taro.CellStyle — Type.

A Cell style. Wrapper for Java class org.apache.poi.ss.usermodel.CellStyle

Taro.CellType

Taro.DAY_MILLISECONDS

Taro.DataFormat

Taro.HOURS_PER_DAY

Taro.MINUTES_PER_HOUR

Taro.ParseOptions

Taro.Row — Type.

A row in a sheet. Contains cells

Taro.SECONDS_PER_DAY

Taro.SECONDS_PER_MINUTE

Taro.Sheet — Type.

An excel Sheet, contained within a workbook. Wrapper around the Java class org.apache.poi.ss.usermodel.Sheet.

Taro.Taro

Taro.Workbook — Type.

An excel Workbook, representing a single file. Wrapper around the Java class org.apache.poi.ss.usermodel.Workbook. Constructors of this types are used to read existing files, or create new ones.

Taro.__init__

Taro.colnum — Function.

Return the integer column number given a character column eg: A -> 1 AA -> 27

Taro.colstring — Function.

Return the character refrence of an integer column number eg. 1 -> A 27 -> AA

Taro.createCell

Taro.createCellStyle — Function.

create a new cell style from a workbook, prior to setting it on a cell

Taro.createRow

Taro.createSheet — Function.

createSheet(w::Workbook, s::AbstractString)

Create a new sheet in the workbook with the specified name.

Taro.eval

Taro.extract — Function.

extract(filename::AbstractString; unsafe = false)

Extract raw text from documents, using Apache Tika. Returns a Dict of metadata name value pairs, and a String with the text of the document.

`filename`: path of file to read. relative to current directory, or absolute
`unsafe` : If set to true, the full contents of the file is read. If false, returned string is capped at 100 000 characters.

Taro.fo — Function.

Taro.fo(inputFoFileName::String, outputPDFFileName::String)

Convert the input fo file to a PDF.

Taro.fop_jar

Taro.fop_lib

Taro.fromExcelDate — Function.

fromExcelDate(date::Number; use1904windowing=false, roundtoSeconds=false)

Convert an Excel style date to a Julia DateTime object.

Excel stores dates and times as a floating point number representing the fractional days since 1/1/1900. If use1904windowing is true, the epoch is 1/1/1904, which is used in some older Excel for Mac versions. If roundtoSeconds is true, the millisecond part of the time is discarded.

Taro.getBooleanCellValue

Taro.getCachedFormulaResultType

Taro.getCell

Taro.getCellFormula — Function.

getCellFormula(cell::Cell)

Taro.getCellType — Function.

getCellType(cell::Cell)

Return the type of a cell: CELLTYPENUMERIC, CELLTYPESTRING, CELLTYPEFORMULA, CELLTYPEBLANK, CELLTYPEBOOLEAN, CELLTYPEERROR

Taro.getCellValue — Function.

getCellValue(cell::Cell)

Return the contents of a Excel cell.

A string or a float value is returned based on the type of the contents of the cell. If a cell is recognised as a being formatted like a date, a Julia DateTime object is returned. This function therefore is not type stable. For formulas, the last evaluated value of the cell is returned.

Note that the dates are stored internally within Excel as floats, and the recognition as a date is heuristic.

If a cell contains an error value, or is empty, nothing is returned.

Taro.getExcelDate — Function.

getExcelDate(date::DateTime, use1904windowing::Bool = false)

Convert a Julia DateTime object into an Excel Date. The result will be a floating point number representing days since 1/1/1900. The time from midnight will be the fractional part of the number. If use1904windowing is true, the epoch is 1/1/1904, which is used in some older Excel for Mac versions.

Taro.getNumericCellValue

Taro.getRow

Taro.getSheet — Function.

getSheet(book::Workbook, sheet)

Return the specified sheet from the workbook. sheet can be specified as a name (string) or number (0-indexed)

Taro.getSheetAt

Taro.getStringCellValue

Taro.include

Taro.init

Taro.isCellDateFormatted — Function.

isCellDateFormatted(cell::Cell)

Return true if the format applied to the cell looks like a date.

This is a heuristic, and not guaranteed to be correct.

Taro.jFile

Taro.readxl — Function.

Read tabular data out of an excel file into a Julia Dataframe. This is similar to the readtable function in the Dataframes package that reads a CSV file into a Dataframe.

The function returns a dataframe from the contents of an MS Excel file. The sheet and region containing the data should be specified. By default, a header row is expected, which must consist only of strings. The header keyword argument should be set to false if no header is present in the data.

filename : path of excel file (.xls or .xlsx)
sheet : sheet name or number (0-based).
    Can be omitted, in which case the first sheet (index `0`) in the workbook is selected.
range : string containing an excel range to read. eg. B4:D45

Optional Arguments : similar to Dataframes.readtable.

header::Bool = true
nastrings::Vector = String["", "NA"]
truestrings::Vector = String["T", "t", "TRUE", "true"]
falsestrings::Vector = String["F", "f", "FALSE", "false"]
colnames::Vector = Symbol[]
coltypes::Vector{Any} = Any[]
skipstart::Int = 0
skiprows::Vector{Int} = Int[]
skipblanks::Bool = true

Taro.setCellFormula — Function.

setCellFormula(c::Cell, formula::AbstractString)

Set a formula as a value to an Excel cell.

The formula string should be what you would expect to enter in excel, but without the +. For example: "A2+2*B2" , "sin(A2)" , "someuserdefined_formula(B2)" Note that the formula will be evaluated only when the file is actually opened in Excel.

Taro.setCellStyle — Function.

setCellStyle(cell:Cell, style::CellStyle)

Set a style to a cell. The CellStyle object must be created from the workbook

Taro.setCellValue — Function.

setCellValue(c::Cell, x)

Set the value of an excel cell. The value can be a string, a real number, or a Date or DateTime.

Taro.setDataFormat — Function.

setDataFormat(w::Workbook, style::CellStyle, format::AbstractString)

Set a dataformat on a CellStyle. Need the workbook to tie everything together

Taro.tika_jar

Taro.writexl — Function.

Write a vector of Julia Dataframes to an Excel file, each representing an Excel Sheet.

filename : path of excel file (.xls or .xlsx)
dfs : A Vector of Table objects with each Table representing a separate Excel Sheet

Optional Arguments.

headers = String[] a string vector of same length as dfs, to add a first line before the sheet is written
sheetnames = String[]
append::Bool = true is supposed to append the DataFrame sheets to an existing excel file, but currently fails often.
    For now, it is better to make sure no such file exists before using.