Docstrings

Parametron.@constraint — Macro.

Add a constraint to the model using operators ==, <=, >=, or in/∈.

in/∈ may only be used for single variables with a right hand side that is one of:

ℤ or Integers

{0, 1} or ZeroOne

Examples

The constraint x >= zeros(2) can be added to a model as follows:

julia> x = [Variable(model) for i = 1 : 2];

julia> @constraint(model, x >= zeros(2))

The constraint that variable x[1] should be an integer can be expressed using:

julia> @constraint(model, x[1] ∈ ℤ)

Parametron.@expression — Macro.

Create a new LazyExpression and apply optimizations to it to reduce allocations and improve performance.

Expressions that do not depend on Parameters or other [LazyExpression]s are simply evaluated straight away.

Examples

Creating an expression that represents p * x1, where p is a parameter that always evaluates to 2:

julia> x1 = Variable(model);

julia> p = Parameter{Int}(() -> 2, model)
Parameter{Int64, …}(…)

julia> expr = @expression p * x1
LazyExpression{FunctionWrapper{…}(LazyExpression{typeof(*), …}(…))}(…)

julia> expr()
2 * x1

Creating an expression that represents p ⋅ x, where p is a parameter that evaluates to [1, 2] and x is a vector of two variables:

julia> model = Parametron.mock_model();

julia> x = Variable.(1 : 2);

julia> p = Parameter(model, val=[1, 2])
Parameter{Array{Int64,1}, …}(…)

julia> expr = @expression p ⋅ x
LazyExpression{FunctionWrapper{…}(LazyExpression{typeof(Parametron.Functions.vecdot!), …}(…))}(…)

julia> expr()
1 * x1 + 2 * x2 + 0

julia> @allocated expr()
0

Note that evaluating the expression does not allocate, because the ⋅ operation is optimized and transformed into a call to the in-place Functions.vecdot! function.

Parametron.@objective — Macro.

Set the objective function of the model.

Examples

Let model be a Model instance. The objective 'minimize x ⋅ x' can be added as follows:

julia> x = [Variable(model) for i = 1 : 2];

julia> @objective model Minimize x ⋅ x;

Parametron.Functions.AffineFunction — Type.

struct AffineFunction{T}

A scalar affine function represented by a sum of LinearTerms and a constant.

AffineFunction overloads the call operator, which can be used to evalute the function given values for the decision variables. The call operator takes an AbstractDict{Variable, T} collection, which associates values with variables.

Example

julia> x, y = Variable.(1 : 2);

julia> affinefun = 2 * x + 3 * y + 4
2 * x1 + 3 * x2 + 4

julia> vals = Dict(x => 4, y => -3);

julia> affinefun(vals)
3

Parametron.Constraint

Parametron.Constraints

Parametron.FunctionWrappersQuickFix

Parametron.Functions — Module.

The Functions module provides types that represent decision variables and functions of these decision variables that possess a certain structure, such as being affine or quadratic in the decision variables.

Various operators are overloaded to make it easy to construct such functions.

Example

julia> x, y = Variable.(1 : 2)
2-element Array{Parametron.Functions.Variable,1}:
 Parametron.Functions.Variable(1)
 Parametron.Functions.Variable(2)

julia> fun = 3 * (x + y) + 4
3 * x1 + 3 * x2 + 4

julia> typeof(fun)
Parametron.Functions.AffineFunction{Int64}

julia> fun.linear
2-element Array{Parametron.Functions.LinearTerm{Int64},1}:

 3 * x1
 3 * x2

julia> fun.constant
Base.RefValue{Int64}(4)

The Functions module also provides in-place versions of certain common operations, e.g., add!, subtract!, and vecdot!, which may be used to evaluate operations performed on the functions into a preallocated destination without any heap allocations.

Parametron.IdentityVarMap

Parametron.LazyExpression — Type.

struct LazyExpression{F, A}

Represents an expression that may be evaluated at a later time, by storing both a function, f, and a tuple of function arguments, args.

LazyExpressions are typically not manually constructed by a user, and hence are not exported. Instead, LazyExpressions should be created using the [@expression] macro.

A LazyExpression may be evaluated by simply calling it with no arguments.

Example

julia> a = ones(2); b = ones(2);

julia> expr = Parametron.LazyExpression(+, a, b)
LazyExpression{typeof(+), …}(…)

julia> expr()
2-element Array{Float64,1}:
 2.0
 2.0

julia> b .= 2
2-element Array{Float64,1}:
 2.0
 2.0

julia> expr()
2-element Array{Float64,1}:
 3.0
 3.0

Parametron.Functions.LinearTerm — Type.

struct LinearTerm{T}

Represents a scalar linear term, i.e. a decision variable scaled by a coefficient.

Parametron.MOI

Parametron.MOIU

Parametron.Maximize

Parametron.Minimize

Parametron.MockModel

Parametron.Model — Type.

Model(optimizer)

Create a new Model, representing an optimization problem to be solved by the optimizer optimizer (a MathOptInterface.AbstractOptimizer).

Parametron.Objective

Parametron.Parameter — Type.

struct Parameter{T, F, InPlace}

Represents a 'placeholder' for data; a value that may be filled in later.

Parameters can be evaluated by simply calling them with no arguments.

Parameters keep track of whether they've already been evaluated using dirty flag. To reevalute a parameter, the dirty flag must first be set using setdirty!(parameter). The update function will then be called when the parameter itself is called.

Examples

julia> value = Ref(1)
Base.RefValue{Int64}(1)

julia> p = Parameter{Int}(() -> value[], model)
Parameter{Int64, …}(…)

julia> p()
1

julia> value[] = 2
2

julia> p()
1

julia> Parametron.setdirty!(p); p()
2

Parametron.Parametron

Parametron.ParametronMOIModel

Parametron.ParametronMOIModelScalarConstraints

Parametron.ParametronMOIModelVectorConstraints

Parametron.Functions.QuadraticFunction — Type.

struct QuadraticFunction{T}

A scalar quadratic function represented by a sum of QuadraticTerms and an AffineFunction.

QuadraticFunction overloads the call operator, which can be used to evalute the function given values for the decision variables. The call operator takes an AbstractDict{Variable, T} collection, which associates values with variables.

Examples

julia> x, y = Variable.(1 : 2);

julia> quadraticfun = 2 * x^2 + 3 * x * y - 2 * y + 4
2 * x1 * x1 + 3 * x1 * x2 + -2 * x2 + 4

julia> vals = Dict(x => 4, y => -3);

julia> quadraticfun(vals)
6

Parametron.Functions.QuadraticTerm — Type.

struct QuadraticTerm{T}

Represents a scalar quadratic term, i.e. the product of two decision variables scaled by a coefficient.

Parametron.Sense

Parametron.Functions.Variable — Type.

struct Variable

Represents a single decision variable.

Parametron.WrappedExpression

Parametron._add_nonnegative_constraint!

Parametron._add_nonpositive_constraint!

Parametron._add_zero_constraint!

Parametron._debug_args

Parametron.add_binary_constraint!

Parametron.add_constraint

Parametron.add_integer_constraint!

Parametron.add_nonnegative_constraint!

Parametron.add_nonpositive_constraint!

Parametron.add_zero_constraint!

Parametron.addparameter!

Parametron.canonical_function_type

Parametron.dualstatus — Function.

dualstatus(m)

Return information regarding the dual of the problem.

Parametron.eval

Parametron.evalarg

Parametron.evalexpr

Parametron.findallocs — Function.

findallocs(x)

Utility function that can be used to track down allocations in LazyExpressions.

Examples

The following session shows the output of findallocs if the expression doesn't allocate:

julia> x = [Variable(model) for i in 1 : 2];

julia> param = Parameter{Int}(() -> 2, model)
Parameter{Int64, …}(…)

julia> expr = @expression param * x
LazyExpression{FunctionWrapper{…}(LazyExpression{typeof(Parametron.Functions.scale!), …}(…))}(…)

julia> Parametron.findallocs(expr)
LazyExpression{FunctionWrapper{…}(LazyExpression{typeof(Parametron.Functions.scale!), …}(…))}(…): 0 bytes
  [1]: Array{LinearTerm{Int64},1}
  [2]: Parameter{Int64, …}(…): 0 bytes
  [3]: Array{Variable,1}

In this session, param allocates, and findallocs reports the allocation:

julia> x = [Variable(model) for i in 1 : 2];

julia> param = Parameter(() -> zeros(2), model)
Parameter{Array{Float64,1}, …}(…)

julia> expr = @expression param ⋅ x
LazyExpression{FunctionWrapper{…}(LazyExpression{typeof(Parametron.Functions.vecdot!), …}(…))}(…)

julia> Parametron.findallocs(expr)
LazyExpression{FunctionWrapper{…}(LazyExpression{typeof(Parametron.Functions.vecdot!), …}(…))}(…): 0 bytes
  [1]: AffineFunction{Float64}
  [2]: Parameter{Array{Float64,1}, …}(…): 96 bytes
  [3]: Array{Variable,1}

Parametron.include

Parametron.initialize! — Function.

initialize!(m)

Copy the problem to be solved to the optimizer.

Users should generally not need to call this function directly, as it is automatically called the first time solve! is called on a Model.

Parametron.isinplace

Parametron.make_moi_equivalent

Parametron.mapindices!

Parametron.mock_model

Parametron.moi_to_native_type

Parametron.objectivevalue — Function.

objectivevalue(m)

Return the value of the objective function at the solution found by the optimizer.

Parametron.optimize

Parametron.optimize_toplevel

Parametron.optimizearg

Parametron.primalstatus — Function.

primalstatus(m)

Return information regarding the primal of the problem.

Parametron.setdirty! — Function.

setdirty!(model)

Mark all parameters associated with the model as 'dirty' (out of date), meaning they must be updated upon their next evaluation.

Parametron.setobjective! — Function.

setobjective!(m, sense, expr)

Set the objective function and optimization sense (Minimize or Maximize).

Parametron.solve! — Function.

solve!(m)

Solve the model m. (Re-)evaluate constraint and objective expressions, update the optimizer's internal representation of the problem, and start the optimization procedure.

Parametron.sort_and_combine! — Function.

sort_and_combine!(v; alg, combine, by, lt)

Sort the vector v in-place using Base.sort!, then combine all entries which compare equal (as determined using the lt and by functions), using the function combined = combine(x, y).

See documentation for Base.sort! regarding the alg, by and lt keyword arguments.

Parametron.terminationstatus — Function.

terminationstatus(m)

Return the termination status of the solver.

Parametron.update! — Function.

Re-evaluate the expressions used to build the constraints and objective function of Model m.

Users should generally not need to call this function directly, as it is automatically called in solve!.

Parametron.value — Function.

value(m, x)

Return the value of variable x as determined by the optimizer.

Parametron.wrap — Function.

wrap(expr)

Wrap a LazyExpression in a FunctionWrappers.FunctionWrapper and return a new LazyExpression with the FunctionWrapper as the function f and an empty tuple as the arguments arg.

The type parameters of the returned LazyFunction depend only on the type of the value returned by expr. This is useful when a common interface is needed for different LazyExpressions that share the same return value type.