Docstrings

NamedTupleTools.@namedtuple

NamedTupleTools.NamedTupleTools — Module.

 NamedTupleTools

This module provides some useful NamedTuple tooling.

see namedtuple, isprototype, fieldnames, fieldtypes, fieldvalues delete, merge

NamedTupleTools.delete — Function.

delete(namedtuple, symbol(s)|Tuple) delete(ntprototype, symbol(s)|Tuple)

Generate a namedtuple [ntprototype] from the first arg omitting fields present in the second arg.

see: merge

NamedTupleTools.eval

NamedTupleTools.fieldvalues — Function.

fieldvalues

obtain values assigned to fields of a struct type (in field order)

NamedTupleTools.include

NamedTupleTools.isprototype — Function.

isprototype( ntprototype )
isprototype( namedtuple  )

Predicate that identifies NamedTuple prototypes.

see: namedtuple

NamedTupleTools.lengthof

NamedTupleTools.namedtuple — Function.

namedtuple(  name1, name2, ..  )
namedtuple( (name1, name2, ..) )
namedtuple(  namedtuple )

Generate a NamedTuple prototype by specifying or obtaining the fieldnames. The prototype is applied to fieldvalues, giving a completed NamedTuple.

Example

julia> ntprototype = namedtuple( :a, :b, :c )

NamedTuple{(:a, :b, :c),T} where T<:Tuple

julia> nt123 = ntprototype(1, 2, 3)

(a = 1, b = 2, c = 3)

julia> ntAb3 = ntprototype("A", "b", 3)

(a = "A", b = "b", c = 3)

see: isprototype

NamedTupleTools.ntfromstruct

NamedTupleTools.structfromnt