Docstrings

ElectronDisplay.CONFIG — Constant.

ElectronDisplay.CONFIG

Configuration for ElectronDisplay.

showable: A callable with signature showable(mime::String, x::Any) :: Bool. This determines if object x is displayed by ElectronDisplay. Default is to return false if mime is "text/html" or "text/markdown" and otherwise fallbacks to Base.showable(mime, x).

single_window::Bool = false: If true, reuse existing window for displaying a new content. If false (default), create a new window for each display.

ElectronDisplay.CachedDataResourceString

ElectronDisplay.DataresourceTableTraitsWrapper

ElectronDisplay.ElectronDisplay

ElectronDisplay.ElectronDisplayConfig

ElectronDisplay.ElectronDisplayType

ElectronDisplay.__init__

ElectronDisplay._display

ElectronDisplay._getglobalwindow

ElectronDisplay._window

ElectronDisplay.asmarkdown

ElectronDisplay.asset

ElectronDisplay.displayhtml

ElectronDisplay.displayhtmlbody

ElectronDisplay.electron_showable

ElectronDisplay.electrondisplay — Function.

electrondisplay([mime,] x)

Show x in Electron window. Use MIME mime if specified.

ElectronDisplay.eval

ElectronDisplay.include