Docstrings

Turing.+

Turing.-

Turing.Core.@VarName — Macro.

Usage: @VarName x[1,2][1+5][45][3] return: (:x,[1,2],6,45,3)

Turing.Core.@model — Macro.

@model(body)

Macro to specify a probabilistic model.

Example:

Model definition:

@model model_generator(x = default_x, y) = begin
    ...
end

Expanded model definition

# Allows passing arguments as kwargs
model_generator(; x = nothing, y = nothing)) = model_generator(x, y)
function model_generator(x = nothing, y = nothing)
    pvars, dvars = Turing.get_vars(Tuple{:x, :y}, (x = x, y = y))
    data = Turing.get_data(dvars, (x = x, y = y))
    defaults = Turing.get_default_values(dvars, (x = default_x, y = nothing))
    
    inner_function(sampler::Turing.AnySampler, model) = inner_function(model)
    function inner_function(model)
        return inner_function(Turing.VarInfo(), Turing.SampleFromPrior(), model)
    end
    function inner_function(vi::Turing.VarInfo, model)
        return inner_function(vi, Turing.SampleFromPrior(), model)
    end
    # Define the main inner function
    function inner_function(vi::Turing.VarInfo, sampler::Turing.AnySampler, model)
        local x
        if isdefined(model.data, :x)
            x = model.data.x
        else
            x = model_defaults.x
        end
        local y
        if isdefined(model.data, :y)
            y = model.data.y
        else
            y = model.defaults.y
        end

        vi.logp = zero(Real)
        ...
    end
    model = Turing.Model{pvars, dvars}(inner_function, data, defaults)
    return model
end

Generating a model: model_generator(x_value)::Model.

Turing.AbstractMixtureModel

Turing.AbstractMvNormal

Turing.AbstractSampler

Turing.Arcsine

Turing.Bernoulli

Turing.Beta

Turing.BetaBinomial

Turing.BetaPrime

Turing.Binomial

Turing.BinomialLogit

Turing.Biweight

Turing.CACHEIDCS

Turing.CACHERANGES

Turing.CACHERESET

Turing.CSMC

Turing.CTask

Turing.Categorical

Turing.Cauchy

Turing.Chains

Turing.Chi

Turing.Chisq

Turing.Continuous

Turing.ContinuousDistribution

Turing.ContinuousMatrixDistribution

Turing.ContinuousMultivariateDistribution

Turing.ContinuousUnivariateDistribution

Turing.Core

Turing.Cosine

Turing.DiagNormal

Turing.DiagNormalCanon

Turing.Dirichlet

Turing.DirichletMultinomial

Turing.Discrete

Turing.DiscreteDistribution

Turing.DiscreteMatrixDistribution

Turing.DiscreteMultivariateDistribution

Turing.DiscreteUniform

Turing.DiscreteUnivariateDistribution

Turing.Distribution

Turing.Distributions

Turing.DoubleExponential

Turing.DynamicHMC

Turing.Inference.DynamicNUTS — Type.

DynamicNUTS(n_iters::Integer)

Dynamic No U-Turn Sampling algorithm provided by the DynamicHMC package. To use it, make sure you have the DynamicHMC package installed.

# Import Turing and DynamicHMC.
using DynamicHMC, Turing

# Model definition.
@model gdemo(x, y) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt(s))
  x ~ Normal(m, sqrt(s))
  y ~ Normal(m, sqrt(s))
  return s, m
end

# Pull 2,000 samples using DynamicNUTS.
chn = sample(gdemo(1.5, 2.0), DynamicNUTS(2000))

Turing.EdgeworthMean

Turing.EdgeworthSum

Turing.EdgeworthZ

Turing.EmpiricalUnivariateDistribution

Turing.Epanechnikov

Turing.Erlang

Turing.Estimator

Turing.Exponential

Turing.FDist

Turing.FisherNoncentralHypergeometric

Turing.Flat

Turing.FlatPos

Turing.Frechet

Turing.FullNormal

Turing.FullNormalCanon

Turing.Gamma

Turing.GeneralizedExtremeValue

Turing.GeneralizedPareto

Turing.Geometric

Turing.Inference.Gibbs — Type.

Gibbs(n_iters, algs...)

Compositional MCMC interface. Gibbs sampling combines one or more sampling algorithms, each of which samples from a different set of variables in a model.

Example:

@model gibbs_example(x) = begin
    v1 ~ Normal(0,1)
    v2 ~ Categorical(5)
        ...
end

# Use PG for a 'v2' variable, and use HMC for the 'v1' variable.
# Note that v2 is discrete, so the PG sampler is more appropriate
# than is HMC.
alg = Gibbs(1000, HMC(1, 0.2, 3, :v1), PG(20, 1, :v2))

Tips:

HMC and NUTS are fast samplers, and can throw off particle-based

methods like Particle Gibbs. You can increase the effectiveness of particle sampling by including more particles in the particle sampler.

Turing.Gumbel

Turing.Inference.HMC — Type.

HMC(n_iters::Int, epsilon::Float64, tau::Int)

Hamiltonian Monte Carlo sampler.

Arguments:

n_iters::Int : The number of samples to pull.

epsilon::Float64 : The leapfrog step size to use.

tau::Int : The number of leapfrop steps to use.

Usage:

HMC(1000, 0.05, 10)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
    s ~ InverseGamma(2,3)
    m ~ Normal(0, sqrt(s))
    x[1] ~ Normal(m, sqrt(s))
    x[2] ~ Normal(m, sqrt(s))
    return s, m
end

sample(gdemo([1.5, 2]), HMC(1000, 0.05, 10))

Tips:

If you are receiving gradient errors when using HMC, try reducing the

step_size parameter.

# Original step_size
sample(gdemo([1.5, 2]), HMC(1000, 0.1, 10))

# Reduced step_size.
sample(gdemo([1.5, 2]), HMC(1000, 0.01, 10))

Turing.Inference.HMCDA — Type.

HMCDA(n_iters::Int, n_adapts::Int, delta::Float64, lambda::Float64)

Hamiltonian Monte Carlo sampler with Dual Averaging algorithm.

Usage:

HMCDA(1000, 200, 0.65, 0.3)

Arguments:

n_iters::Int : Number of samples to pull.

n_adapts::Int : Numbers of samples to use for adaptation.

delta::Float64 : Target acceptance rate. 65% is often recommended.

lambda::Float64 : Target leapfrop length.

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0, sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), HMCDA(1000, 200, 0.65, 0.3))

For more information, please view the following paper (arXiv link):

Hoffman, Matthew D., and Andrew Gelman. "The No-U-turn sampler: adaptively setting path lengths in Hamiltonian Monte Carlo." Journal of Machine Learning Research 15, no. 1 (2014): 1593-1623.

Turing.Hypergeometric

Turing.Inference.IPMCMC — Type.

IPMCMC(n_particles::Int, n_iters::Int, n_nodes::Int, n_csmc_nodes::Int)

Particle Gibbs sampler.

Note that this method is particle-based, and arrays of variables must be stored in a TArray object.

Usage:

IPMCMC(100, 100, 4, 2)

Arguments:

n_particles::Int : Number of particles to use.

n_iters::Int : Number of iterations to employ.

n_nodes::Int : The number of nodes running SMC and CSMC.

n_csmc_nodes::Int : The number of CSMC nodes.

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), IPMCMC(100, 100, 4, 2))

A paper on this can be found here.

Turing.Inference.IS — Type.

IS(n_particles::Int)

Importance sampling algorithm.

Note that this method is particle-based, and arrays of variables must be stored in a TArray object.

Arguments:

n_particles is the number of particles to use.

Usage:

IS(1000)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
    s ~ InverseGamma(2,3)
    m ~ Normal(0,sqrt.(s))
    x[1] ~ Normal(m, sqrt.(s))
    x[2] ~ Normal(m, sqrt.(s))
    return s, m
end

sample(gdemo([1.5, 2]), IS(1000))

Turing.Inference

Turing.InverseGamma

Turing.InverseGaussian

Turing.InverseWishart

Turing.IsoNormal

Turing.IsoNormalCanon

Turing.KSDist

Turing.KSOneSided

Turing.Kolmogorov

Turing.Laplace

Turing.Levy

Turing.Libtask

Turing.LocationScale

Turing.LogDensityProblems

Turing.LogNormal

Turing.Logistic

Turing.MCMCChain

Turing.Inference.MH — Type.

MH(n_iters::Int)

Metropolis-Hastings sampler.

Usage:

MH(100, (:m, (x) -> Normal(x, 0.1)))

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

chn = sample(gdemo([1.5, 2]), MH(1000))

Turing.MLEstimator

Turing.MatrixDistribution

Turing.Matrixvariate

Turing.MixtureModel

Turing.Model — Type.

struct Model{pvars, dvars, F, TData, TDefaults}
    f::F
    data::TData
    defaults::TDefaults
end

A Model struct with parameter variables pvars, data variables dvars, inner function f, data::NamedTuple and defaults::NamedTuple.

Turing.Multinomial

Turing.Multivariate

Turing.MultivariateDistribution

Turing.MultivariateMixture

Turing.MultivariateNormal

Turing.MvLogNormal

Turing.MvNormal

Turing.MvNormalCanon

Turing.MvNormalKnownCov

Turing.MvTDist

Turing.Inference.NUTS — Type.

NUTS(n_iters::Int, n_adapts::Int, delta::Float64)

No-U-Turn Sampler (NUTS) sampler.

Usage:

NUTS(1000, 200, 0.6j_max)

Arguments:

n_iters::Int : The number of samples to pull.

n_adapts::Int : The number of samples to use with adapatation.

delta::Float64 : Target acceptance rate.

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0, sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.j_max, 2]), NUTS(1000, 200, 0.6j_max))

Turing.NegativeBinomial

Turing.NonMatrixDistribution

Turing.NoncentralBeta

Turing.NoncentralChisq

Turing.NoncentralF

Turing.NoncentralHypergeometric

Turing.NoncentralT

Turing.Normal

Turing.NormalCanon

Turing.NormalInverseGaussian

Turing.Inference.PG — Type.

PG(n_particles::Int, n_iters::Int)

Particle Gibbs sampler.

Note that this method is particle-based, and arrays of variables must be stored in a TArray object.

Usage:

PG(100, 100)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0, sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), PG(100, 100))

Turing.PIMH

Turing.Inference.PMMH — Type.

PMMH(n_iters::Int, smc_alg:::SMC, parameters_algs::Tuple{MH})

Particle independant Metropolis–Hastings and Particle marginal Metropolis–Hastings samplers.

Note that this method is particle-based, and arrays of variables must be stored in a TArray object.

Usage:

alg = PMMH(100, SMC(20, :v1), MH(1,:v2))
alg = PMMH(100, SMC(20, :v1), MH(1,(:v2, (x) -> Normal(x, 1))))

Arguments:

n_iters::Int : Number of iterations to run.

smc_alg:::SMC : An SMC algorithm to use.

parameters_algs::Tuple{MH} : An MH algorithm, which includes a

sample space specification.

Turing.PROGRESS

Turing.Pareto

Turing.Poisson

Turing.PoissonBinomial

Turing.QQPair

Turing.Rayleigh

Turing.RealInterval

Turing.Inference.SGHMC — Type.

SGHMC(n_iters::Int, learning_rate::Float64, momentum_decay::Float64)

Stochastic Gradient Hamiltonian Monte Carlo sampler.

Usage:

SGHMC(1000, 0.01, 0.1)

Arguments:

n_iters::Int : Number of samples to pull.

learning_rate::Float64 : The learning rate.

momentum_decay::Float64 : Momentum decay variable.

Example:

@model example begin
  ...
end

sample(example, SGHMC(1000, 0.01, 0.1))

Turing.Inference.SGLD — Type.

SGLD(n_iters::Int, epsilon::Float64)

Stochastic Gradient Langevin Dynamics sampler.

Usage:

SGLD(1000, 0.5)

Arguments:

n_iters::Int : Number of samples to pull.

epsilon::Float64 : The scaling factor for the learing rate.

Example:

@model example begin
  ...
end

sample(example, SGLD(1000, 0.5))

Turing.Inference.SMC — Type.

SMC(n_particles::Int)

Sequential Monte Carlo sampler.

Note that this method is particle-based, and arrays of variables must be stored in a TArray object.

Usage:

SMC(1000)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0, sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), SMC(1000))

Turing.Sampleable

Turing.Sampler — Type.

Sampler{T}

Generic interface for implementing inference algorithms. An implementation of an algorithm should include the following:

A type specifying the algorithm and its parameters, derived from InferenceAlgorithm

A method of sample function that produces results of inference, which is where actual inference happens.

Turing translates models to chunks that call the modelling functions at specified points. The dispatch is based on the value of a sampler variable. To include a new inference algorithm implements the requirements mentioned above in a separate file, then include that file at the end of this one.

Turing.Semicircle

Turing.Skellam

Turing.SufficientStats

Turing.SymTriangularDist

Turing.TArray

Turing.TDist

Turing.TRef

Turing.TriangularDist

Turing.Triweight

Turing.Truncated

Turing.TruncatedNormal

Turing.Turing

Turing.Uniform

Turing.Univariate

Turing.UnivariateDistribution

Turing.UnivariateGMM

Turing.UnivariateMixture

Turing.Utilities

Turing.ValueSupport

Turing.VariateForm

Turing.VecBinomialLogit

Turing.VonMises

Turing.VonMisesFisher

Turing.WalleniusNoncentralHypergeometric

Turing.Weibull

Turing.Wishart

Turing.ZeroMeanDiagNormal

Turing.ZeroMeanDiagNormalCanon

Turing.ZeroMeanFullNormal

Turing.ZeroMeanFullNormalCanon

Turing.ZeroMeanIsoNormal

Turing.ZeroMeanIsoNormalCanon

Turing.__init__

Turing.__inits__

Turing.autcorplot

Turing.auto_tune_chunk_size!

Turing.autocorplot

Turing.autocorplot!

Turing.binaryentropy

Turing.canonform

Turing.ccdf

Turing.cdf

Turing.cf

Turing.cgf

Turing.circvar

Turing.component

Turing.components

Turing.componentwise_logpdf

Turing.componentwise_pdf

Turing.concentration

Turing.consume

Turing.cor

Turing.corner

Turing.corner!

Turing.cov

Turing.cquantile

Turing.cumulant

Turing.densityplot

Turing.describe

Turing.dim

Turing.discretediag

Turing.dof

Turing.enable_stack_copying

Turing.entropy

Turing.estimate

Turing.eval

Turing.expected_logdet

Turing.failprob

Turing.fit

Turing.fit_map

Turing.fit_map!

Turing.fit_mle

Turing.fit_mle!

Turing.freecumulant

Turing.gelmandiag

Turing.get

Turing.get_defaults

Turing.get_dvars

Turing.get_pvars

Turing.getindex

Turing.gewekediag

Turing.gradlogpdf

Turing.hasfinitesupport

Turing.heideldiag

Turing.histogramplot

Turing.in_dvars

Turing.in_pvars

Turing.include

Turing.insupport

Turing.invcov

Turing.invlogccdf

Turing.invlogcdf

Turing.invscale

Turing.isbounded

Turing.isleptokurtic

Turing.islowerbounded

Turing.ismesokurtic

Turing.isplatykurtic

Turing.isprobvec

Turing.isupperbounded

Turing.kde

Turing.kurtosis

Turing.location

Turing.location!

Turing.logccdf

Turing.logcdf

Turing.logdetcov

Turing.loglikelihood

Turing.logpdf

Turing.logpdf!

Turing.mean

Turing.meandir

Turing.meanform

Turing.meanlogx

Turing.meanplot

Turing.meanplot!

Turing.median

Turing.mgf

Turing.mixeddensity

Turing.mixeddensity!

Turing.mixeddensityplot

Turing.mode

Turing.modes

Turing.moment

Turing.ncategories

Turing.ncomponents

Turing.nsamples

Turing.ntrials

Turing.params

Turing.params!

Turing.partype

Turing.pdf

Turing.plot

Turing.probs

Turing.probval

Turing.produce

Turing.qqbuild

Turing.quantile

Turing.rafterydiag

Turing.rate

Turing.resume

Turing.runmodel!

Turing.sample

Turing.sample!

Turing.sampler

Turing.scale

Turing.scale!

Turing.setadbackend

Turing.setadsafe

Turing.setchunksize

Turing.setindex!

Turing.shape

Turing.skewness

Turing.span

Turing.sqmahal

Turing.sqmahal!

Turing.std

Turing.stdlogx

Turing.succprob

Turing.suffstats

Turing.support

Turing.test_distr

Turing.test_samples

Turing.tfill

Turing.traceplot

Turing.traceplot!

Turing.turnprogress

Turing.tzeros

Turing.var

Turing.varlogx

Turing.wsample

Turing.wsample!