Docstrings

MbedTLS.@debug

MbedTLS.@err_check

MbedTLS.@lockdata

MbedTLS.@💀

MbedTLS.@😬

MbedTLS.@🤖

MbedTLS.CIPHER_3DES

MbedTLS.CIPHER_AES

MbedTLS.CIPHER_AES_128_CBC

MbedTLS.CIPHER_AES_128_CCM

MbedTLS.CIPHER_AES_128_CFB128

MbedTLS.CIPHER_AES_128_CTR

MbedTLS.CIPHER_AES_128_ECB

MbedTLS.CIPHER_AES_128_GCM

MbedTLS.CIPHER_AES_192_CBC

MbedTLS.CIPHER_AES_192_CCM

MbedTLS.CIPHER_AES_192_CFB128

MbedTLS.CIPHER_AES_192_CTR

MbedTLS.CIPHER_AES_192_ECB

MbedTLS.CIPHER_AES_192_GCM

MbedTLS.CIPHER_AES_256_CBC

MbedTLS.CIPHER_AES_256_CCM

MbedTLS.CIPHER_AES_256_CFB128

MbedTLS.CIPHER_AES_256_CTR

MbedTLS.CIPHER_AES_256_ECB

MbedTLS.CIPHER_AES_256_GCM

MbedTLS.CIPHER_ARC4

MbedTLS.CIPHER_ARC4_128

MbedTLS.CIPHER_BLOWFISH

MbedTLS.CIPHER_BLOWFISH_CBC

MbedTLS.CIPHER_BLOWFISH_CFB64

MbedTLS.CIPHER_BLOWFISH_CTR

MbedTLS.CIPHER_BLOWFISH_ECB

MbedTLS.CIPHER_CAMELLIA

MbedTLS.CIPHER_CAMELLIA_128_CBC

MbedTLS.CIPHER_CAMELLIA_128_CCM

MbedTLS.CIPHER_CAMELLIA_128_CFB128

MbedTLS.CIPHER_CAMELLIA_128_CTR

MbedTLS.CIPHER_CAMELLIA_128_ECB

MbedTLS.CIPHER_CAMELLIA_128_GCM

MbedTLS.CIPHER_CAMELLIA_192_CBC

MbedTLS.CIPHER_CAMELLIA_192_CCM

MbedTLS.CIPHER_CAMELLIA_192_CFB128

MbedTLS.CIPHER_CAMELLIA_192_CTR

MbedTLS.CIPHER_CAMELLIA_192_ECB

MbedTLS.CIPHER_CAMELLIA_192_GCM

MbedTLS.CIPHER_CAMELLIA_256_CBC

MbedTLS.CIPHER_CAMELLIA_256_CCM

MbedTLS.CIPHER_CAMELLIA_256_CFB128

MbedTLS.CIPHER_CAMELLIA_256_CTR

MbedTLS.CIPHER_CAMELLIA_256_ECB

MbedTLS.CIPHER_CAMELLIA_256_GCM

MbedTLS.CIPHER_DES

MbedTLS.CIPHER_DES_CBC

MbedTLS.CIPHER_DES_ECB

MbedTLS.CIPHER_DES_EDE3_CBC

MbedTLS.CIPHER_DES_EDE3_ECB

MbedTLS.CIPHER_DES_EDE_CBC

MbedTLS.CIPHER_DES_EDE_ECB

MbedTLS.CIPHER_ID_NONE

MbedTLS.CIPHER_ID_NULL

MbedTLS.CIPHER_MODE_CBC

MbedTLS.CIPHER_MODE_CCM

MbedTLS.CIPHER_MODE_CFB

MbedTLS.CIPHER_MODE_CTR

MbedTLS.CIPHER_MODE_ECB

MbedTLS.CIPHER_MODE_GCM

MbedTLS.CIPHER_MODE_NONE

MbedTLS.CIPHER_MODE_OFB

MbedTLS.CIPHER_MODE_STREAM

MbedTLS.CIPHER_NONE

MbedTLS.CIPHER_NULL

MbedTLS.CRT

MbedTLS.Cipher — Type.

Cipher(info::Union{CipherID, CipherKind}) -> Cipher

Construct a cipher object and set it to use the specified cipher algorithm.

The algorithm can either be specific (ie, CIPHER_AES_256_CBC), or general (ie, CIPHER_AES). In the latter case, a default choice of specific cipher will be used. See ?CipherInfo for more details.

MbedTLS.CipherID

MbedTLS.CipherInfo — Type.

CipherInfo(id::CipherID, key_bitlength, mode::CipherMode) -> CipherInfo

Construct a custom cipher info object.

id: A CipherID, such as CIPHERAES of CIPHERBLOWFIHS

key_bitlength: The bit length of the secret key. Available options depend

on the specific cipher id.

mode: Either Encrypt or Decrypt to indicate which cipher operation will

be performed with this cipher info.

CipherInfo(id::CipherID) -> CipherInfo

Build a CipherInfo for the given cipher id using the strongest available key size and the CBC block mode.

Warning: In CBC block mode, it is imperative that you use a unique IV (initial value) for each encryption operation to maintain security.

MbedTLS.CipherKind

MbedTLS.CipherMode

MbedTLS.CtrDrbg

MbedTLS.DEBUG_LEVEL

MbedTLS.DECRYPT

MbedTLS.DebugThreshold

MbedTLS.ENCRYPT

MbedTLS.ERROR

MbedTLS.Entropy

MbedTLS.INFO

MbedTLS.MBEDTLS_AESNI_AES

MbedTLS.MBEDTLS_AESNI_CLMUL

MbedTLS.MBEDTLS_AES_DECRYPT

MbedTLS.MBEDTLS_AES_ENCRYPT

MbedTLS.MBEDTLS_ASN1_BIT_STRING

MbedTLS.MBEDTLS_ASN1_BMP_STRING

MbedTLS.MBEDTLS_ASN1_BOOLEAN

MbedTLS.MBEDTLS_ASN1_CONSTRUCTED

MbedTLS.MBEDTLS_ASN1_CONTEXT_SPECIFIC

MbedTLS.MBEDTLS_ASN1_GENERALIZED_TIME

MbedTLS.MBEDTLS_ASN1_IA5_STRING

MbedTLS.MBEDTLS_ASN1_INTEGER

MbedTLS.MBEDTLS_ASN1_NULL

MbedTLS.MBEDTLS_ASN1_OCTET_STRING

MbedTLS.MBEDTLS_ASN1_OID

MbedTLS.MBEDTLS_ASN1_PRIMITIVE

MbedTLS.MBEDTLS_ASN1_PRINTABLE_STRING

MbedTLS.MBEDTLS_ASN1_SEQUENCE

MbedTLS.MBEDTLS_ASN1_SET

MbedTLS.MBEDTLS_ASN1_T61_STRING

MbedTLS.MBEDTLS_ASN1_UNIVERSAL_STRING

MbedTLS.MBEDTLS_ASN1_UTC_TIME

MbedTLS.MBEDTLS_ASN1_UTF8_STRING

MbedTLS.MBEDTLS_BLOWFISH_BLOCKSIZE

MbedTLS.MBEDTLS_BLOWFISH_DECRYPT

MbedTLS.MBEDTLS_BLOWFISH_ENCRYPT

MbedTLS.MBEDTLS_BLOWFISH_MAX_KEY_BITS

MbedTLS.MBEDTLS_BLOWFISH_MIN_KEY_BITS

MbedTLS.MBEDTLS_BLOWFISH_ROUNDS

MbedTLS.MBEDTLS_CAMELLIA_DECRYPT

MbedTLS.MBEDTLS_CAMELLIA_ENCRYPT

MbedTLS.MBEDTLS_CIPHERSUITE_NODTLS

MbedTLS.MBEDTLS_CIPHERSUITE_SHORT_TAG

MbedTLS.MBEDTLS_CIPHERSUITE_WEAK

MbedTLS.MBEDTLS_CIPHER_VARIABLE_IV_LEN

MbedTLS.MBEDTLS_CIPHER_VARIABLE_KEY_LEN

MbedTLS.MBEDTLS_CTR_DRBG_BLOCKSIZE

MbedTLS.MBEDTLS_CTR_DRBG_ENTROPY_LEN

MbedTLS.MBEDTLS_CTR_DRBG_KEYSIZE

MbedTLS.MBEDTLS_CTR_DRBG_MAX_INPUT

MbedTLS.MBEDTLS_CTR_DRBG_MAX_REQUEST

MbedTLS.MBEDTLS_CTR_DRBG_MAX_SEED_INPUT

MbedTLS.MBEDTLS_CTR_DRBG_PR_OFF

MbedTLS.MBEDTLS_CTR_DRBG_PR_ON

MbedTLS.MBEDTLS_CTR_DRBG_RESEED_INTERVAL

MbedTLS.MBEDTLS_DES_DECRYPT

MbedTLS.MBEDTLS_DES_ENCRYPT

MbedTLS.MBEDTLS_DES_KEY_SIZE

MbedTLS.MBEDTLS_ECP_DP_MAX

MbedTLS.MBEDTLS_ECP_FIXED_POINT_OPTIM

MbedTLS.MBEDTLS_ECP_MAX_BITS

MbedTLS.MBEDTLS_ECP_PF_COMPRESSED

MbedTLS.MBEDTLS_ECP_PF_UNCOMPRESSED

MbedTLS.MBEDTLS_ECP_TLS_NAMED_CURVE

MbedTLS.MBEDTLS_ECP_WINDOW_SIZE

MbedTLS.MBEDTLS_ENTROPY_BLOCK_SIZE

MbedTLS.MBEDTLS_ENTROPY_MAX_GATHER

MbedTLS.MBEDTLS_ENTROPY_MAX_SEED_SIZE

MbedTLS.MBEDTLS_ENTROPY_MAX_SOURCES

MbedTLS.MBEDTLS_ENTROPY_MIN_HARDCLOCK

MbedTLS.MBEDTLS_ENTROPY_MIN_HARDWARE

MbedTLS.MBEDTLS_ENTROPY_MIN_HAVEGE

MbedTLS.MBEDTLS_ENTROPY_MIN_PLATFORM

MbedTLS.MBEDTLS_ENTROPY_SOURCE_STRONG

MbedTLS.MBEDTLS_ENTROPY_SOURCE_WEAK

MbedTLS.MBEDTLS_ERR_AES_INVALID_INPUT_LENGTH

MbedTLS.MBEDTLS_ERR_AES_INVALID_KEY_LENGTH

MbedTLS.MBEDTLS_ERR_ASN1_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_ASN1_BUF_TOO_SMALL

MbedTLS.MBEDTLS_ERR_ASN1_INVALID_DATA

MbedTLS.MBEDTLS_ERR_ASN1_INVALID_LENGTH

MbedTLS.MBEDTLS_ERR_ASN1_LENGTH_MISMATCH

MbedTLS.MBEDTLS_ERR_ASN1_OUT_OF_DATA

MbedTLS.MBEDTLS_ERR_ASN1_UNEXPECTED_TAG

MbedTLS.MBEDTLS_ERR_BASE64_BUFFER_TOO_SMALL

MbedTLS.MBEDTLS_ERR_BASE64_INVALID_CHARACTER

MbedTLS.MBEDTLS_ERR_BLOWFISH_INVALID_INPUT_LENGTH

MbedTLS.MBEDTLS_ERR_BLOWFISH_INVALID_KEY_LENGTH

MbedTLS.MBEDTLS_ERR_CAMELLIA_INVALID_INPUT_LENGTH

MbedTLS.MBEDTLS_ERR_CAMELLIA_INVALID_KEY_LENGTH

MbedTLS.MBEDTLS_ERR_CCM_AUTH_FAILED

MbedTLS.MBEDTLS_ERR_CCM_BAD_INPUT

MbedTLS.MBEDTLS_ERR_CIPHER_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_CIPHER_AUTH_FAILED

MbedTLS.MBEDTLS_ERR_CIPHER_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_CIPHER_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_CIPHER_FULL_BLOCK_EXPECTED

MbedTLS.MBEDTLS_ERR_CIPHER_INVALID_PADDING

MbedTLS.MBEDTLS_ERR_CTR_DRBG_ENTROPY_SOURCE_FAILED

MbedTLS.MBEDTLS_ERR_CTR_DRBG_FILE_IO_ERROR

MbedTLS.MBEDTLS_ERR_CTR_DRBG_INPUT_TOO_BIG

MbedTLS.MBEDTLS_ERR_CTR_DRBG_REQUEST_TOO_BIG

MbedTLS.MBEDTLS_ERR_DES_INVALID_INPUT_LENGTH

MbedTLS.MBEDTLS_ERR_DHM_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_DHM_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_DHM_CALC_SECRET_FAILED

MbedTLS.MBEDTLS_ERR_DHM_FILE_IO_ERROR

MbedTLS.MBEDTLS_ERR_DHM_INVALID_FORMAT

MbedTLS.MBEDTLS_ERR_DHM_MAKE_PARAMS_FAILED

MbedTLS.MBEDTLS_ERR_DHM_MAKE_PUBLIC_FAILED

MbedTLS.MBEDTLS_ERR_DHM_READ_PARAMS_FAILED

MbedTLS.MBEDTLS_ERR_DHM_READ_PUBLIC_FAILED

MbedTLS.MBEDTLS_ERR_ECP_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_ECP_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_ECP_BUFFER_TOO_SMALL

MbedTLS.MBEDTLS_ERR_ECP_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_ECP_INVALID_KEY

MbedTLS.MBEDTLS_ERR_ECP_RANDOM_FAILED

MbedTLS.MBEDTLS_ERR_ECP_SIG_LEN_MISMATCH

MbedTLS.MBEDTLS_ERR_ECP_VERIFY_FAILED

MbedTLS.MBEDTLS_ERR_ENTROPY_FILE_IO_ERROR

MbedTLS.MBEDTLS_ERR_ENTROPY_MAX_SOURCES

MbedTLS.MBEDTLS_ERR_ENTROPY_NO_SOURCES_DEFINED

MbedTLS.MBEDTLS_ERR_ENTROPY_NO_STRONG_SOURCE

MbedTLS.MBEDTLS_ERR_ENTROPY_SOURCE_FAILED

MbedTLS.MBEDTLS_ERR_GCM_AUTH_FAILED

MbedTLS.MBEDTLS_ERR_GCM_BAD_INPUT

MbedTLS.MBEDTLS_ERR_HMAC_DRBG_ENTROPY_SOURCE_FAILED

MbedTLS.MBEDTLS_ERR_HMAC_DRBG_FILE_IO_ERROR

MbedTLS.MBEDTLS_ERR_HMAC_DRBG_INPUT_TOO_BIG

MbedTLS.MBEDTLS_ERR_HMAC_DRBG_REQUEST_TOO_BIG

MbedTLS.MBEDTLS_ERR_MD_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_MD_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_MD_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_MD_FILE_IO_ERROR

MbedTLS.MBEDTLS_ERR_MPI_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_MPI_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_MPI_BUFFER_TOO_SMALL

MbedTLS.MBEDTLS_ERR_MPI_DIVISION_BY_ZERO

MbedTLS.MBEDTLS_ERR_MPI_FILE_IO_ERROR

MbedTLS.MBEDTLS_ERR_MPI_INVALID_CHARACTER

MbedTLS.MBEDTLS_ERR_MPI_NEGATIVE_VALUE

MbedTLS.MBEDTLS_ERR_MPI_NOT_ACCEPTABLE

MbedTLS.MBEDTLS_ERR_NET_ACCEPT_FAILED

MbedTLS.MBEDTLS_ERR_NET_BIND_FAILED

MbedTLS.MBEDTLS_ERR_NET_BUFFER_TOO_SMALL

MbedTLS.MBEDTLS_ERR_NET_CONNECT_FAILED

MbedTLS.MBEDTLS_ERR_NET_CONN_RESET

MbedTLS.MBEDTLS_ERR_NET_INVALID_CONTEXT

MbedTLS.MBEDTLS_ERR_NET_LISTEN_FAILED

MbedTLS.MBEDTLS_ERR_NET_RECV_FAILED

MbedTLS.MBEDTLS_ERR_NET_SEND_FAILED

MbedTLS.MBEDTLS_ERR_NET_SOCKET_FAILED

MbedTLS.MBEDTLS_ERR_NET_UNKNOWN_HOST

MbedTLS.MBEDTLS_ERR_OID_BUF_TOO_SMALL

MbedTLS.MBEDTLS_ERR_OID_NOT_FOUND

MbedTLS.MBEDTLS_ERR_PADLOCK_DATA_MISALIGNED

MbedTLS.MBEDTLS_ERR_PEM_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_PEM_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_PEM_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_PEM_INVALID_DATA

MbedTLS.MBEDTLS_ERR_PEM_INVALID_ENC_IV

MbedTLS.MBEDTLS_ERR_PEM_NO_HEADER_FOOTER_PRESENT

MbedTLS.MBEDTLS_ERR_PEM_PASSWORD_MISMATCH

MbedTLS.MBEDTLS_ERR_PEM_PASSWORD_REQUIRED

MbedTLS.MBEDTLS_ERR_PEM_UNKNOWN_ENC_ALG

MbedTLS.MBEDTLS_ERR_PKCS12_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_PKCS12_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_PKCS12_PASSWORD_MISMATCH

MbedTLS.MBEDTLS_ERR_PKCS12_PBE_INVALID_FORMAT

MbedTLS.MBEDTLS_ERR_PKCS5_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_PKCS5_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_PKCS5_INVALID_FORMAT

MbedTLS.MBEDTLS_ERR_PKCS5_PASSWORD_MISMATCH

MbedTLS.MBEDTLS_ERR_PK_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_PK_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_PK_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_PK_FILE_IO_ERROR

MbedTLS.MBEDTLS_ERR_PK_INVALID_ALG

MbedTLS.MBEDTLS_ERR_PK_INVALID_PUBKEY

MbedTLS.MBEDTLS_ERR_PK_KEY_INVALID_FORMAT

MbedTLS.MBEDTLS_ERR_PK_KEY_INVALID_VERSION

MbedTLS.MBEDTLS_ERR_PK_PASSWORD_MISMATCH

MbedTLS.MBEDTLS_ERR_PK_PASSWORD_REQUIRED

MbedTLS.MBEDTLS_ERR_PK_SIG_LEN_MISMATCH

MbedTLS.MBEDTLS_ERR_PK_TYPE_MISMATCH

MbedTLS.MBEDTLS_ERR_PK_UNKNOWN_NAMED_CURVE

MbedTLS.MBEDTLS_ERR_PK_UNKNOWN_PK_ALG

MbedTLS.MBEDTLS_ERR_RSA_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_RSA_INVALID_PADDING

MbedTLS.MBEDTLS_ERR_RSA_KEY_CHECK_FAILED

MbedTLS.MBEDTLS_ERR_RSA_KEY_GEN_FAILED

MbedTLS.MBEDTLS_ERR_RSA_OUTPUT_TOO_LARGE

MbedTLS.MBEDTLS_ERR_RSA_PRIVATE_FAILED

MbedTLS.MBEDTLS_ERR_RSA_PUBLIC_FAILED

MbedTLS.MBEDTLS_ERR_RSA_RNG_FAILED

MbedTLS.MBEDTLS_ERR_RSA_VERIFY_FAILED

MbedTLS.MBEDTLS_ERR_SSL_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_CERTIFICATE

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_CERTIFICATE_REQUEST

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_CERTIFICATE_VERIFY

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_CHANGE_CIPHER_SPEC

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_CLIENT_HELLO

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_CLIENT_KEY_EXCHANGE

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_CLIENT_KEY_EXCHANGE_CS

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_CLIENT_KEY_EXCHANGE_RP

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_FINISHED

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_NEW_SESSION_TICKET

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_PROTOCOL_VERSION

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_SERVER_HELLO

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_SERVER_HELLO_DONE

MbedTLS.MBEDTLS_ERR_SSL_BAD_HS_SERVER_KEY_EXCHANGE

MbedTLS.MBEDTLS_ERR_SSL_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_SSL_BUFFER_TOO_SMALL

MbedTLS.MBEDTLS_ERR_SSL_CA_CHAIN_REQUIRED

MbedTLS.MBEDTLS_ERR_SSL_CERTIFICATE_REQUIRED

MbedTLS.MBEDTLS_ERR_SSL_CERTIFICATE_TOO_LARGE

MbedTLS.MBEDTLS_ERR_SSL_COMPRESSION_FAILED

MbedTLS.MBEDTLS_ERR_SSL_CONN_EOF

MbedTLS.MBEDTLS_ERR_SSL_COUNTER_WRAPPING

MbedTLS.MBEDTLS_ERR_SSL_FATAL_ALERT_MESSAGE

MbedTLS.MBEDTLS_ERR_SSL_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_SSL_HELLO_VERIFY_REQUIRED

MbedTLS.MBEDTLS_ERR_SSL_HW_ACCEL_FAILED

MbedTLS.MBEDTLS_ERR_SSL_HW_ACCEL_FALLTHROUGH

MbedTLS.MBEDTLS_ERR_SSL_INTERNAL_ERROR

MbedTLS.MBEDTLS_ERR_SSL_INVALID_MAC

MbedTLS.MBEDTLS_ERR_SSL_INVALID_RECORD

MbedTLS.MBEDTLS_ERR_SSL_NO_CIPHER_CHOSEN

MbedTLS.MBEDTLS_ERR_SSL_NO_CLIENT_CERTIFICATE

MbedTLS.MBEDTLS_ERR_SSL_NO_RNG

MbedTLS.MBEDTLS_ERR_SSL_NO_USABLE_CIPHERSUITE

MbedTLS.MBEDTLS_ERR_SSL_PEER_CLOSE_NOTIFY

MbedTLS.MBEDTLS_ERR_SSL_PEER_VERIFY_FAILED

MbedTLS.MBEDTLS_ERR_SSL_PK_TYPE_MISMATCH

MbedTLS.MBEDTLS_ERR_SSL_PRIVATE_KEY_REQUIRED

MbedTLS.MBEDTLS_ERR_SSL_SESSION_TICKET_EXPIRED

MbedTLS.MBEDTLS_ERR_SSL_TIMEOUT

MbedTLS.MBEDTLS_ERR_SSL_UNEXPECTED_MESSAGE

MbedTLS.MBEDTLS_ERR_SSL_UNKNOWN_CIPHER

MbedTLS.MBEDTLS_ERR_SSL_UNKNOWN_IDENTITY

MbedTLS.MBEDTLS_ERR_SSL_WAITING_SERVER_HELLO_RENEGO

MbedTLS.MBEDTLS_ERR_SSL_WANT_READ

MbedTLS.MBEDTLS_ERR_SSL_WANT_WRITE

MbedTLS.MBEDTLS_ERR_THREADING_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_THREADING_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_THREADING_MUTEX_ERROR

MbedTLS.MBEDTLS_ERR_X509_ALLOC_FAILED

MbedTLS.MBEDTLS_ERR_X509_BAD_INPUT_DATA

MbedTLS.MBEDTLS_ERR_X509_BUFFER_TOO_SMALL

MbedTLS.MBEDTLS_ERR_X509_CERT_UNKNOWN_FORMAT

MbedTLS.MBEDTLS_ERR_X509_CERT_VERIFY_FAILED

MbedTLS.MBEDTLS_ERR_X509_FEATURE_UNAVAILABLE

MbedTLS.MBEDTLS_ERR_X509_FILE_IO_ERROR

MbedTLS.MBEDTLS_ERR_X509_INVALID_ALG

MbedTLS.MBEDTLS_ERR_X509_INVALID_DATE

MbedTLS.MBEDTLS_ERR_X509_INVALID_EXTENSIONS

MbedTLS.MBEDTLS_ERR_X509_INVALID_FORMAT

MbedTLS.MBEDTLS_ERR_X509_INVALID_NAME

MbedTLS.MBEDTLS_ERR_X509_INVALID_SERIAL

MbedTLS.MBEDTLS_ERR_X509_INVALID_SIGNATURE

MbedTLS.MBEDTLS_ERR_X509_INVALID_VERSION

MbedTLS.MBEDTLS_ERR_X509_SIG_MISMATCH

MbedTLS.MBEDTLS_ERR_X509_UNKNOWN_OID

MbedTLS.MBEDTLS_ERR_X509_UNKNOWN_SIG_ALG

MbedTLS.MBEDTLS_ERR_X509_UNKNOWN_VERSION

MbedTLS.MBEDTLS_ERR_XTEA_INVALID_INPUT_LENGTH

MbedTLS.MBEDTLS_GCM_DECRYPT

MbedTLS.MBEDTLS_GCM_ENCRYPT

MbedTLS.MBEDTLS_HAVEGE_COLLECT_SIZE

MbedTLS.MBEDTLS_HMAC_DRBG_MAX_INPUT

MbedTLS.MBEDTLS_HMAC_DRBG_MAX_REQUEST

MbedTLS.MBEDTLS_HMAC_DRBG_MAX_SEED_INPUT

MbedTLS.MBEDTLS_HMAC_DRBG_PR_OFF

MbedTLS.MBEDTLS_HMAC_DRBG_PR_ON

MbedTLS.MBEDTLS_HMAC_DRBG_RESEED_INTERVAL

MbedTLS.MBEDTLS_LN_2_DIV_LN_10_SCALE100

MbedTLS.MBEDTLS_MAX_BLOCK_LENGTH

MbedTLS.MBEDTLS_MAX_IV_LENGTH

MbedTLS.MBEDTLS_MD_MAX_SIZE

MbedTLS.MBEDTLS_MEMORY_ALIGN_MULTIPLE

MbedTLS.MBEDTLS_MEMORY_VERIFY_NONE

MbedTLS.MBEDTLS_MPI_MAX_LIMBS

MbedTLS.MBEDTLS_MPI_MAX_SIZE

MbedTLS.MBEDTLS_MPI_WINDOW_SIZE

MbedTLS.MBEDTLS_NET_LISTEN_BACKLOG

MbedTLS.MBEDTLS_NET_PROTO_TCP

MbedTLS.MBEDTLS_NET_PROTO_UDP

MbedTLS.MBEDTLS_PADLOCK_ACE

MbedTLS.MBEDTLS_PADLOCK_PHE

MbedTLS.MBEDTLS_PADLOCK_PMM

MbedTLS.MBEDTLS_PADLOCK_RNG

MbedTLS.MBEDTLS_PKCS12_DERIVE_IV

MbedTLS.MBEDTLS_PKCS12_DERIVE_KEY

MbedTLS.MBEDTLS_PKCS12_DERIVE_MAC_KEY

MbedTLS.MBEDTLS_PKCS12_PBE_DECRYPT

MbedTLS.MBEDTLS_PKCS12_PBE_ENCRYPT

MbedTLS.MBEDTLS_PKCS5_DECRYPT

MbedTLS.MBEDTLS_PKCS5_ENCRYPT

MbedTLS.MBEDTLS_PK_DEBUG_MAX_ITEMS

MbedTLS.MBEDTLS_PSK_MAX_LEN

MbedTLS.MBEDTLS_RSA_CRYPT

MbedTLS.MBEDTLS_RSA_PKCS_V15

MbedTLS.MBEDTLS_RSA_PKCS_V21

MbedTLS.MBEDTLS_RSA_PRIVATE

MbedTLS.MBEDTLS_RSA_PUBLIC

MbedTLS.MBEDTLS_RSA_SALT_LEN_ANY

MbedTLS.MBEDTLS_RSA_SIGN

MbedTLS.MBEDTLS_SSL_ALERT_LEVEL_FATAL

MbedTLS.MBEDTLS_SSL_ALERT_LEVEL_WARNING

MbedTLS.MBEDTLS_SSL_ALERT_MSG_ACCESS_DENIED

MbedTLS.MBEDTLS_SSL_ALERT_MSG_BAD_CERT

MbedTLS.MBEDTLS_SSL_ALERT_MSG_BAD_RECORD_MAC

MbedTLS.MBEDTLS_SSL_ALERT_MSG_CERT_EXPIRED

MbedTLS.MBEDTLS_SSL_ALERT_MSG_CERT_REVOKED

MbedTLS.MBEDTLS_SSL_ALERT_MSG_CERT_UNKNOWN

MbedTLS.MBEDTLS_SSL_ALERT_MSG_CLOSE_NOTIFY

MbedTLS.MBEDTLS_SSL_ALERT_MSG_DECODE_ERROR

MbedTLS.MBEDTLS_SSL_ALERT_MSG_DECOMPRESSION_FAILURE

MbedTLS.MBEDTLS_SSL_ALERT_MSG_DECRYPTION_FAILED

MbedTLS.MBEDTLS_SSL_ALERT_MSG_DECRYPT_ERROR

MbedTLS.MBEDTLS_SSL_ALERT_MSG_EXPORT_RESTRICTION

MbedTLS.MBEDTLS_SSL_ALERT_MSG_HANDSHAKE_FAILURE

MbedTLS.MBEDTLS_SSL_ALERT_MSG_ILLEGAL_PARAMETER

MbedTLS.MBEDTLS_SSL_ALERT_MSG_INAPROPRIATE_FALLBACK

MbedTLS.MBEDTLS_SSL_ALERT_MSG_INSUFFICIENT_SECURITY

MbedTLS.MBEDTLS_SSL_ALERT_MSG_INTERNAL_ERROR

MbedTLS.MBEDTLS_SSL_ALERT_MSG_NO_APPLICATION_PROTOCOL

MbedTLS.MBEDTLS_SSL_ALERT_MSG_NO_CERT

MbedTLS.MBEDTLS_SSL_ALERT_MSG_NO_RENEGOTIATION

MbedTLS.MBEDTLS_SSL_ALERT_MSG_PROTOCOL_VERSION

MbedTLS.MBEDTLS_SSL_ALERT_MSG_RECORD_OVERFLOW

MbedTLS.MBEDTLS_SSL_ALERT_MSG_UNEXPECTED_MESSAGE

MbedTLS.MBEDTLS_SSL_ALERT_MSG_UNKNOWN_CA

MbedTLS.MBEDTLS_SSL_ALERT_MSG_UNKNOWN_PSK_IDENTITY

MbedTLS.MBEDTLS_SSL_ALERT_MSG_UNRECOGNIZED_NAME

MbedTLS.MBEDTLS_SSL_ALERT_MSG_UNSUPPORTED_CERT

MbedTLS.MBEDTLS_SSL_ALERT_MSG_UNSUPPORTED_EXT

MbedTLS.MBEDTLS_SSL_ALERT_MSG_USER_CANCELED

MbedTLS.MBEDTLS_SSL_ANTI_REPLAY_DISABLED

MbedTLS.MBEDTLS_SSL_ANTI_REPLAY_ENABLED

MbedTLS.MBEDTLS_SSL_ARC4_DISABLED

MbedTLS.MBEDTLS_SSL_ARC4_ENABLED

MbedTLS.MBEDTLS_SSL_CACHE_DEFAULT_MAX_ENTRIES

MbedTLS.MBEDTLS_SSL_CACHE_DEFAULT_TIMEOUT

MbedTLS.MBEDTLS_SSL_CBC_RECORD_SPLITTING_DISABLED

MbedTLS.MBEDTLS_SSL_CBC_RECORD_SPLITTING_ENABLED

MbedTLS.MBEDTLS_SSL_CERT_TYPE_ECDSA_SIGN

MbedTLS.MBEDTLS_SSL_CERT_TYPE_RSA_SIGN

MbedTLS.MBEDTLS_SSL_CHANNEL_INBOUND

MbedTLS.MBEDTLS_SSL_CHANNEL_OUTBOUND

MbedTLS.MBEDTLS_SSL_COMPRESSION_ADD

MbedTLS.MBEDTLS_SSL_COMPRESS_DEFLATE

MbedTLS.MBEDTLS_SSL_COMPRESS_NULL

MbedTLS.MBEDTLS_SSL_COOKIE_TIMEOUT

MbedTLS.MBEDTLS_SSL_DEFAULT_TICKET_LIFETIME

MbedTLS.MBEDTLS_SSL_DTLS_TIMEOUT_DFL_MAX

MbedTLS.MBEDTLS_SSL_DTLS_TIMEOUT_DFL_MIN

MbedTLS.MBEDTLS_SSL_EMPTY_RENEGOTIATION_INFO

MbedTLS.MBEDTLS_SSL_ETM_DISABLED

MbedTLS.MBEDTLS_SSL_ETM_ENABLED

MbedTLS.MBEDTLS_SSL_EXTENDED_MS_DISABLED

MbedTLS.MBEDTLS_SSL_EXTENDED_MS_ENABLED

MbedTLS.MBEDTLS_SSL_FALLBACK_SCSV_VALUE

MbedTLS.MBEDTLS_SSL_HASH_MD5

MbedTLS.MBEDTLS_SSL_HASH_NONE

MbedTLS.MBEDTLS_SSL_HASH_SHA1

MbedTLS.MBEDTLS_SSL_HASH_SHA224

MbedTLS.MBEDTLS_SSL_HASH_SHA256

MbedTLS.MBEDTLS_SSL_HASH_SHA384

MbedTLS.MBEDTLS_SSL_HASH_SHA512

MbedTLS.MBEDTLS_SSL_HS_CERTIFICATE

MbedTLS.MBEDTLS_SSL_HS_CERTIFICATE_REQUEST

MbedTLS.MBEDTLS_SSL_HS_CERTIFICATE_VERIFY

MbedTLS.MBEDTLS_SSL_HS_CLIENT_HELLO

MbedTLS.MBEDTLS_SSL_HS_CLIENT_KEY_EXCHANGE

MbedTLS.MBEDTLS_SSL_HS_FINISHED

MbedTLS.MBEDTLS_SSL_HS_HELLO_REQUEST

MbedTLS.MBEDTLS_SSL_HS_HELLO_VERIFY_REQUEST

MbedTLS.MBEDTLS_SSL_HS_NEW_SESSION_TICKET

MbedTLS.MBEDTLS_SSL_HS_SERVER_HELLO

MbedTLS.MBEDTLS_SSL_HS_SERVER_HELLO_DONE

MbedTLS.MBEDTLS_SSL_HS_SERVER_KEY_EXCHANGE

MbedTLS.MBEDTLS_SSL_INITIAL_HANDSHAKE

MbedTLS.MBEDTLS_SSL_IS_CLIENT

MbedTLS.MBEDTLS_SSL_IS_FALLBACK

MbedTLS.MBEDTLS_SSL_IS_NOT_FALLBACK

MbedTLS.MBEDTLS_SSL_IS_SERVER

MbedTLS.MBEDTLS_SSL_LEGACY_ALLOW_RENEGOTIATION

MbedTLS.MBEDTLS_SSL_LEGACY_BREAK_HANDSHAKE

MbedTLS.MBEDTLS_SSL_LEGACY_NO_RENEGOTIATION

MbedTLS.MBEDTLS_SSL_LEGACY_RENEGOTIATION

MbedTLS.MBEDTLS_SSL_MAC_ADD

MbedTLS.MBEDTLS_SSL_MAJOR_VERSION_3

MbedTLS.MBEDTLS_SSL_MAX_CONTENT_LEN

MbedTLS.MBEDTLS_SSL_MAX_FRAG_LEN_1024

MbedTLS.MBEDTLS_SSL_MAX_FRAG_LEN_2048

MbedTLS.MBEDTLS_SSL_MAX_FRAG_LEN_4096

MbedTLS.MBEDTLS_SSL_MAX_FRAG_LEN_512

MbedTLS.MBEDTLS_SSL_MAX_FRAG_LEN_INVALID

MbedTLS.MBEDTLS_SSL_MAX_FRAG_LEN_NONE

MbedTLS.MBEDTLS_SSL_MINOR_VERSION_0

MbedTLS.MBEDTLS_SSL_MINOR_VERSION_1

MbedTLS.MBEDTLS_SSL_MINOR_VERSION_2

MbedTLS.MBEDTLS_SSL_MINOR_VERSION_3

MbedTLS.MBEDTLS_SSL_MSG_ALERT

MbedTLS.MBEDTLS_SSL_MSG_APPLICATION_DATA

MbedTLS.MBEDTLS_SSL_MSG_CHANGE_CIPHER_SPEC

MbedTLS.MBEDTLS_SSL_MSG_HANDSHAKE

MbedTLS.MBEDTLS_SSL_PADDING_ADD

MbedTLS.MBEDTLS_SSL_PRESET_DEFAULT

MbedTLS.MBEDTLS_SSL_PRESET_SUITEB

MbedTLS.MBEDTLS_SSL_RENEGOTIATION_DISABLED

MbedTLS.MBEDTLS_SSL_RENEGOTIATION_DONE

MbedTLS.MBEDTLS_SSL_RENEGOTIATION_ENABLED

MbedTLS.MBEDTLS_SSL_RENEGOTIATION_IN_PROGRESS

MbedTLS.MBEDTLS_SSL_RENEGOTIATION_NOT_ENFORCED

MbedTLS.MBEDTLS_SSL_RENEGOTIATION_PENDING

MbedTLS.MBEDTLS_SSL_RENEGO_MAX_RECORDS_DEFAULT

MbedTLS.MBEDTLS_SSL_RETRANS_FINISHED

MbedTLS.MBEDTLS_SSL_RETRANS_PREPARING

MbedTLS.MBEDTLS_SSL_RETRANS_SENDING

MbedTLS.MBEDTLS_SSL_RETRANS_WAITING

MbedTLS.MBEDTLS_SSL_SECURE_RENEGOTIATION

MbedTLS.MBEDTLS_SSL_SESSION_TICKETS_DISABLED

MbedTLS.MBEDTLS_SSL_SESSION_TICKETS_ENABLED

MbedTLS.MBEDTLS_SSL_SIG_ANON

MbedTLS.MBEDTLS_SSL_SIG_ECDSA

MbedTLS.MBEDTLS_SSL_SIG_RSA

MbedTLS.MBEDTLS_SSL_TRANSPORT_DATAGRAM

MbedTLS.MBEDTLS_SSL_TRANSPORT_STREAM

MbedTLS.MBEDTLS_SSL_TRUNCATED_HMAC_LEN

MbedTLS.MBEDTLS_SSL_TRUNC_HMAC_DISABLED

MbedTLS.MBEDTLS_SSL_TRUNC_HMAC_ENABLED

MbedTLS.MBEDTLS_SSL_VERIFY_DATA_MAX_LEN

MbedTLS.MBEDTLS_SSL_VERIFY_NONE

MbedTLS.MBEDTLS_SSL_VERIFY_OPTIONAL

MbedTLS.MBEDTLS_SSL_VERIFY_REQUIRED

MbedTLS.MBEDTLS_SSL_VERIFY_UNSET

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_128_CCM

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_128_CCM_8

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_256_CCM

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_256_CCM_8

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_AES_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_CAMELLIA_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_CAMELLIA_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_CAMELLIA_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_NULL_SHA

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_NULL_SHA256

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_NULL_SHA384

MbedTLS.MBEDTLS_TLS_DHE_PSK_WITH_RC4_128_SHA

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_128_CCM

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_128_CCM_8

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_256_CBC_SHA256

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_256_CCM

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_256_CCM_8

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_AES_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_CAMELLIA_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_CAMELLIA_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_CAMELLIA_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_CAMELLIA_256_CBC_SHA256

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_CAMELLIA_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_DHE_RSA_WITH_DES_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_128_CCM

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_128_CCM_8

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_256_CCM

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_256_CCM_8

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_CAMELLIA_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_CAMELLIA_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_CAMELLIA_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_NULL_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_ECDSA_WITH_RC4_128_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_AES_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_CAMELLIA_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_NULL_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_NULL_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_NULL_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_PSK_WITH_RC4_128_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_CAMELLIA_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_CAMELLIA_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_CAMELLIA_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_NULL_SHA

MbedTLS.MBEDTLS_TLS_ECDHE_RSA_WITH_RC4_128_SHA

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_AES_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_AES_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_AES_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_CAMELLIA_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_CAMELLIA_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_CAMELLIA_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_NULL_SHA

MbedTLS.MBEDTLS_TLS_ECDH_ECDSA_WITH_RC4_128_SHA

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_AES_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_AES_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_AES_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_CAMELLIA_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_CAMELLIA_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_CAMELLIA_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_NULL_SHA

MbedTLS.MBEDTLS_TLS_ECDH_RSA_WITH_RC4_128_SHA

MbedTLS.MBEDTLS_TLS_EXT_ALPN

MbedTLS.MBEDTLS_TLS_EXT_ENCRYPT_THEN_MAC

MbedTLS.MBEDTLS_TLS_EXT_EXTENDED_MASTER_SECRET

MbedTLS.MBEDTLS_TLS_EXT_MAX_FRAGMENT_LENGTH

MbedTLS.MBEDTLS_TLS_EXT_RENEGOTIATION_INFO

MbedTLS.MBEDTLS_TLS_EXT_SERVERNAME

MbedTLS.MBEDTLS_TLS_EXT_SERVERNAME_HOSTNAME

MbedTLS.MBEDTLS_TLS_EXT_SESSION_TICKET

MbedTLS.MBEDTLS_TLS_EXT_SIG_ALG

MbedTLS.MBEDTLS_TLS_EXT_SUPPORTED_ELLIPTIC_CURVES

MbedTLS.MBEDTLS_TLS_EXT_SUPPORTED_POINT_FORMATS

MbedTLS.MBEDTLS_TLS_EXT_TRUNCATED_HMAC

MbedTLS.MBEDTLS_TLS_PSK_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_128_CCM

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_128_CCM_8

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_256_CCM

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_256_CCM_8

MbedTLS.MBEDTLS_TLS_PSK_WITH_AES_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_PSK_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_PSK_WITH_CAMELLIA_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_PSK_WITH_CAMELLIA_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_PSK_WITH_CAMELLIA_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_PSK_WITH_NULL_SHA

MbedTLS.MBEDTLS_TLS_PSK_WITH_NULL_SHA256

MbedTLS.MBEDTLS_TLS_PSK_WITH_NULL_SHA384

MbedTLS.MBEDTLS_TLS_PSK_WITH_RC4_128_SHA

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_AES_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_AES_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_AES_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_CAMELLIA_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_CAMELLIA_256_CBC_SHA384

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_CAMELLIA_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_NULL_SHA

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_NULL_SHA256

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_NULL_SHA384

MbedTLS.MBEDTLS_TLS_RSA_PSK_WITH_RC4_128_SHA

MbedTLS.MBEDTLS_TLS_RSA_WITH_3DES_EDE_CBC_SHA

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_128_CCM

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_128_CCM_8

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_256_CBC_SHA256

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_256_CCM

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_256_CCM_8

MbedTLS.MBEDTLS_TLS_RSA_WITH_AES_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_RSA_WITH_CAMELLIA_128_CBC_SHA

MbedTLS.MBEDTLS_TLS_RSA_WITH_CAMELLIA_128_CBC_SHA256

MbedTLS.MBEDTLS_TLS_RSA_WITH_CAMELLIA_128_GCM_SHA256

MbedTLS.MBEDTLS_TLS_RSA_WITH_CAMELLIA_256_CBC_SHA

MbedTLS.MBEDTLS_TLS_RSA_WITH_CAMELLIA_256_CBC_SHA256

MbedTLS.MBEDTLS_TLS_RSA_WITH_CAMELLIA_256_GCM_SHA384

MbedTLS.MBEDTLS_TLS_RSA_WITH_DES_CBC_SHA

MbedTLS.MBEDTLS_TLS_RSA_WITH_NULL_MD5

MbedTLS.MBEDTLS_TLS_RSA_WITH_NULL_SHA

MbedTLS.MBEDTLS_TLS_RSA_WITH_NULL_SHA256

MbedTLS.MBEDTLS_TLS_RSA_WITH_RC4_128_MD5

MbedTLS.MBEDTLS_TLS_RSA_WITH_RC4_128_SHA

MbedTLS.MBEDTLS_VERSION_MAJOR

MbedTLS.MBEDTLS_VERSION_MINOR

MbedTLS.MBEDTLS_VERSION_NUMBER

MbedTLS.MBEDTLS_VERSION_PATCH

MbedTLS.MBEDTLS_X509_BADCERT_BAD_KEY

MbedTLS.MBEDTLS_X509_BADCERT_BAD_MD

MbedTLS.MBEDTLS_X509_BADCERT_BAD_PK

MbedTLS.MBEDTLS_X509_BADCERT_CN_MISMATCH

MbedTLS.MBEDTLS_X509_BADCERT_EXPIRED

MbedTLS.MBEDTLS_X509_BADCERT_EXT_KEY_USAGE

MbedTLS.MBEDTLS_X509_BADCERT_FUTURE

MbedTLS.MBEDTLS_X509_BADCERT_KEY_USAGE

MbedTLS.MBEDTLS_X509_BADCERT_MISSING

MbedTLS.MBEDTLS_X509_BADCERT_NOT_TRUSTED

MbedTLS.MBEDTLS_X509_BADCERT_NS_CERT_TYPE

MbedTLS.MBEDTLS_X509_BADCERT_OTHER

MbedTLS.MBEDTLS_X509_BADCERT_REVOKED

MbedTLS.MBEDTLS_X509_BADCERT_SKIP_VERIFY

MbedTLS.MBEDTLS_X509_BADCRL_BAD_KEY

MbedTLS.MBEDTLS_X509_BADCRL_BAD_MD

MbedTLS.MBEDTLS_X509_BADCRL_BAD_PK

MbedTLS.MBEDTLS_X509_BADCRL_EXPIRED

MbedTLS.MBEDTLS_X509_BADCRL_FUTURE

MbedTLS.MBEDTLS_X509_BADCRL_NOT_TRUSTED

MbedTLS.MBEDTLS_X509_CRT_VERSION_1

MbedTLS.MBEDTLS_X509_CRT_VERSION_2

MbedTLS.MBEDTLS_X509_CRT_VERSION_3

MbedTLS.MBEDTLS_X509_FORMAT_DER

MbedTLS.MBEDTLS_X509_FORMAT_PEM

MbedTLS.MBEDTLS_X509_MAX_DN_NAME_SIZE

MbedTLS.MBEDTLS_X509_MAX_INTERMEDIATE_CA

MbedTLS.MBEDTLS_X509_RFC5280_MAX_SERIAL_LEN

MbedTLS.MBEDTLS_X509_RFC5280_UTC_TIME_LEN

MbedTLS.MBEDTLS_XTEA_DECRYPT

MbedTLS.MBEDTLS_XTEA_ENCRYPT

MbedTLS.MBED_SUCCESS

MbedTLS.MD

MbedTLS.MDInfo

MbedTLS.MDKind

MbedTLS.MD_MD2

MbedTLS.MD_MD4

MbedTLS.MD_MD5

MbedTLS.MD_NONE

MbedTLS.MD_SHA

MbedTLS.MD_SHA1

MbedTLS.MD_SHA224

MbedTLS.MD_SHA256

MbedTLS.MD_SHA384

MbedTLS.MbedException

MbedTLS.MbedTLS

MbedTLS.NONE

MbedTLS.OPERATION_NONE

MbedTLS.Operation

MbedTLS.PADDING_NONE

MbedTLS.PADDING_ONE_AND_ZEROS

MbedTLS.PADDING_PKCS7

MbedTLS.PADDING_ZEROS

MbedTLS.PADDING_ZEROS_AND_LEN

MbedTLS.PKContext

MbedTLS.Padding

MbedTLS.RSA

MbedTLS.SSLConfig

MbedTLS.SSLContext

MbedTLS.STATE_CHANGE

MbedTLS.VERBOSE

MbedTLS._CRT_SECURE_NO_DEPRECATE

MbedTLS.__ctr_drbg__init__

MbedTLS.__entropyinit__

MbedTLS.__init__

MbedTLS.__sslinit__

MbedTLS._write

MbedTLS.add_source!

MbedTLS.alpn_proto

MbedTLS.associate!

MbedTLS.authmode!

MbedTLS.bitlength

MbedTLS.c_dbg

MbedTLS.c_entropy

MbedTLS.c_recv

MbedTLS.c_rng

MbedTLS.c_send

MbedTLS.ca_chain!

MbedTLS.check_async_exception

MbedTLS.check_deps

MbedTLS.config_defaults!

MbedTLS.crt_info

MbedTLS.crt_parse

MbedTLS.crt_parse!

MbedTLS.crt_parse_file

MbedTLS.crypt

MbedTLS.crypt!

MbedTLS.dbg!

MbedTLS.debug_header

MbedTLS.decrypt — Function.

decrypt(cipher, key, msg, [iv]) -> Vector{UInt8}

Decrypt a message using the given cipher. The cipher can be specified as

a generic cipher (like CIPHER_AES)

a specific cipher (like CIPHERAES256_CBC)

a Cipher object

key is the symmetric key used for cryptography, given as either a String or a Vector{UInt8}. It must be the right length for the chosen cipher; for example, CIPHERAES256_CBC requires a 32-byte (256-bit) key.

msg is the message to be encoded. It should either be convertible to a String or be a Vector{UInt8}.

iv is the initialization vector, whose size must match the block size of the cipher (eg, 16 bytes for AES) and correspond to the iv used by the encryptor. By default, it will be set to all zeros.

MbedTLS.decrypt!

MbedTLS.depsjl_path

MbedTLS.digest — Function.

digest(kind::MDKind, msg::Vector{UInt8}, [key::Vector{UInt8}]) -> Vector{UInt8}

Perform a digest of the given type on the given message (a byte array), return a byte array with the digest.

If an optional key is given, perform an HMAC digest.

MbedTLS.digest! — Function.

digest!(kind::MDKind, msg::Vector{UInt8}, [key::Vector{UInt8}, ], buffer::Vector{UInt8})

In-place version of digest that stores the digest to buffer.

It is the user's responsibility to ensure that buffer is long enough to contain the digest. get_size(kind::MDKind) returns the appropriate size.

MbedTLS.encrypt — Function.

encrypt(cipher, key, msg, [iv]) -> Vector{UInt8}

Encrypt a message using the given cipher. The cipher can be specified as

a generic cipher (like CIPHER_AES)

a specific cipher (like CIPHERAES256_CBC)

a Cipher object

key is the symmetric key used for cryptography, given as either a String or a Vector{UInt8}. It must be the right length for the chosen cipher; for example, CIPHERAES256_CBC requires a 32-byte (256-bit) key.

msg is the message to be encoded. It should either be convertible to a String or be a Vector{UInt8}.

iv is the initialization vector, whose size must match the block size of the cipher (eg, 16 bytes for AES). By default, it will be set to all zeros, which is not secure. For security reasons, it should be set to a different value for each encryption operation.

MbedTLS.encrypt!

MbedTLS.eval

MbedTLS.f_dbg

MbedTLS.f_recv — Function.

Copy at most nbytes of encrypted data to buf from the bio connection. If no encrypted bytes are available return:

MBEDTLS_ERR_SSL_WANT_READ if the connection is still open, or

MBEDTLS_ERR_NET_CONN_RESET if it is closed.

MbedTLS.f_rng

MbedTLS.f_send — Function.

Copy nbytes of encrypted data from buf to the underlying bio connection.

MbedTLS.finish!

MbedTLS.gather

MbedTLS.gen_key

MbedTLS.gen_key!

MbedTLS.get_ciphersuite

MbedTLS.get_key_bitlen

MbedTLS.get_name

MbedTLS.get_peer_cert

MbedTLS.get_size — Function.

get_size(kind::MDKind) -> Int

Returns the size of the digest in bytes that the given digest type requires.

For example,

get_size(MD_SHA256) == 32

MbedTLS.get_version

MbedTLS.handshake

MbedTLS.handshake!

MbedTLS.hostname!

MbedTLS.include

MbedTLS.jl_entropy

MbedTLS.libmbedcrypto

MbedTLS.libmbedtls

MbedTLS.libmbedx509

MbedTLS.mbed_err

MbedTLS.mbedtls_mpi

MbedTLS.mbedtls_rsa_context

MbedTLS.mpi_import!

MbedTLS.mpi_size

MbedTLS.own_cert!

MbedTLS.parse_key!

MbedTLS.parse_keyfile

MbedTLS.parse_keyfile!

MbedTLS.parse_public_key!

MbedTLS.parse_public_keyfile

MbedTLS.parse_public_keyfile!

MbedTLS.private

MbedTLS.process_iv

MbedTLS.pubkey_from_vals!

MbedTLS.public

MbedTLS.reset!

MbedTLS.rng!

MbedTLS.secrets_log

MbedTLS.secrets_log_file

MbedTLS.secrets_log_state

MbedTLS.seed!

MbedTLS.set_alpn!

MbedTLS.set_bio! — Function.

Connect f_send and f_recv callback functions to SSLContext.

MbedTLS.set_dbg_level

MbedTLS.set_iv!

MbedTLS.set_key!

MbedTLS.set_padding_mode!

MbedTLS.setup!

MbedTLS.sign

MbedTLS.sign!

MbedTLS.ssl_abandon — Function.

The documentation for ssl_read, ssl_write and ssl_close_notify all say:

If this function returns something other than 0 or MBEDTLSERRSSLWANTREAD/WRITE, you must stop using the SSL context for reading or writing, and either free it or call

This function ensures that the SSLContext is won't be used again.

MbedTLS.ssl_check_pending — Function.

ssl_check_pending(::SSLContext)::Bool

Check if there is data already read from the underlying transport but not yet processed.

If the SSL/TLS module successfully returns from an operation - e.g. a handshake or an application record read - and you're awaiting incoming data next, you must not immediately idle on the underlying transport to have data ready, but you need to check the value of this function first. The reason is that the desired data might already be read but not yet processed. If, in contrast, a previous call to the SSL/TLS module returned MBEDTLSERRSSLWANTREAD, it is not necessary to call this function, as the latter error code entails that all internal data has been processed.

https://tls.mbed.org/api/ssl_8h.html#a4075f7de9877fd667bcfa2e819e33426

MbedTLS.ssl_close_notify — Function.

ssl_close_notify(::SSLContext)

Notify the peer that the connection is being closed. 0 if successful, or a specific SSL error code.

If this function returns something other than 0 or MBEDTLSERRSSLWANTREAD/WRITE, you must stop using the SSL context for reading or writing, and either free it or call mbedtlssslsession_reset() on it before re-using it for a new connection; the current connection must be closed.

https://tls.mbed.org/api/ssl_8h.html#ac2c1b17128ead2df3082e27b603deb4c

MbedTLS.ssl_conf_renegotiation! — Function.

Enable / Disable renegotiation support for connection when initiated by peer (MBEDTLSSSLRENEGOTIATIONENABLED or MBEDTLSSSLRENEGOTIATIONDISABLED). See: https://tls.mbed.org/api/ssl_8h.html#aad4f50fc1c0a018fd5eb18fd9621d0d3

MbedTLS.ssl_get_bytes_avail — Function.

Return the number of application data bytes remaining to be read from the current record.

https://tls.mbed.org/api/ssl_8h.html#ad43142085f3182e9b0dc967ec582032b:

MbedTLS.ssl_handshake

MbedTLS.ssl_read — Function.

ssl_read(::SSLContext, ptr, n)

Returns One of the following: 0 if the read end of the underlying transport was closed, the (positive) number of bytes read, or a negative error code on failure.

If MBEDTLSERRSSLWANTREAD is returned, no application data is available from the underlying transport. In this case, the function needs to be called again at some later stage.

If this function returns something other than a positive value or MBEDTLSERRSSLWANTREAD/WRITE or MBEDTLSERRSSLCLIENTRECONNECT, you must stop using the SSL context for reading or writing, and either free it or call mbedtlssslsession_reset() on it before re-using it for a new connection; the current connection must be closed.

https://tls.mbed.org/api/ssl_8h.html#aa2c29eeb1deaf5ad9f01a7515006ede5

MbedTLS.ssl_session_reset — Function.

Reset an already initialized SSL context for re-use while retaining application-set variables, function pointers and data.

MbedTLS.ssl_set_bio

MbedTLS.ssl_setup

MbedTLS.ssl_unsafe_read — Function.

ssl_unsafe_read(::SSLContext, buf, nbytes)

Copies at most nbytes of decrypted data into buf. Never blocks to wait for more data to arrive. Returns number of bytes copied into buf (<= nbytes). Updates ctx.bytesavailable with the number of decrypted bytes remaining in the TLS buffer.

Stops when:

nbytes have been copied, or

there are no more decrypted bytes available in the TLS buffer, or

a TLS close_notify message is received.

When TLS close_notify is received:

isreadable is set to false and bytesavailable is set to zero. [RFC5246 7.2.1]: "Any data received after a closure alert is ignored."

the number of bytes read before the close_notify is returned as usual.

Throws a MbedException if ssl_read returns an unhandled error code.

When an unhandled exception occurs isreadable is set to false.

MbedTLS.ssl_unsafe_write — Function.

Copy nbytes from buf to ctx.

The TLS library function ssl_write is called as many times as needed to send all the data. The TLS library encrypts the data and passes it to the f_send function which sends it to the underlying connection (ctx.bio). See f_send and set_bio! below.

MbedTLS.ssl_write — Function.

ssl_write(::SSLContext, ptr, n)

When this function returns MBEDTLSERRSSLWANTWRITE/READ, it must be called later with the same arguments, until it returns a value greater that or equal to 0. When the function returns MBEDTLSERRSSLWANTWRITE there may be some partial data in the output buffer, however this is not yet sent.

If this function returns something other than 0, a positive value or MBEDTLSERRSSLWANTREAD/WRITE, you must stop using the SSL context for reading or writing, and either free it or call mbedtlssslsession_reset() on it before re-using it for a new connection; the current connection must be closed.

https://tls.mbed.org/api/ssl_8h.html#a5bbda87d484de82df730758b475f32e5

MbedTLS.strerror

MbedTLS.taskid

MbedTLS.tls_dbg

MbedTLS.tls_dbg_log_secrets

MbedTLS.tobytes

MbedTLS.update! — Function.

update!(cipher::Cipher, buf_in, buf_out::Vector{UInt8}) -> Int

Run the given cipher on buf_in (a String or Vector{UInt8}) and store the result of the cipher in buf_out (a Vector{UInt8}).

It is your responsibility to ensure that buf_out is at least as large as necessary to hold the result. It should be at least as big as the size of buf_in plus the block size associated with cipher.

MbedTLS.update_manual

MbedTLS.verify

MbedTLS.wait_for_decrypted_data — Function.

While there are no decrypted bytes available but the connection is readable:

If the TLS buffer has no pending (unprocessed) data wait for more encrypted data to arrive on the underlying connection.

Run a zero-byte read to allow the library to process its internal buffer, and/or read from the underlying connection.

ssl_unsafe_read updates the isreadable and bytesavailable state.

MbedTLS.wait_for_encrypted_data