Docstrings

QuantumOptics.bases.@samebases — Macro.

@samebases

Macro to skip checks for same bases. Useful for *, expect and similar functions.

QuantumOptics.timeevolution.@skiptimechecks — Macro.

@skiptimechecks

Macro to skip checks during time-dependent problems. Useful for timeevolution.master_dynamic and similar functions.

QuantumOptics.operators.AbstractOperator — Type.

Abstract base class for all operators.

All deriving operator classes have to define the fields basis_l and basis_r defining the left and right side bases.

For fast time evolution also at least the function gemv!(alpha, op::AbstractOperator, x::Ket, beta, result::Ket) should be implemented. Many other generic multiplication functions can be defined in terms of this function and are provided automatically.

QuantumOptics.bases.Basis — Type.

Abstract base class for all specialized bases.

The Basis class is meant to specify a basis of the Hilbert space of the studied system. Besides basis specific information all subclasses must implement a shape variable which indicates the dimension of the used Hilbert space. For a spin-1/2 Hilbert space this would be the vector Int[2]. A system composed of two spins would then have a shape vector Int[2 2].

Composite systems can be defined with help of the CompositeBasis class.

QuantumOptics.states.Bra — Type.

Bra(b::Basis[, data])

Bra state defined by coefficients in respect to the basis.

QuantumOptics.bases.CompositeBasis — Type.

CompositeBasis(b1, b2...)

Basis for composite Hilbert spaces.

Stores the subbases in a vector and creates the shape vector directly from the shape vectors of these subbases. Instead of creating a CompositeBasis directly tensor(b1, b2...) or b1 ⊗ b2 ⊗ … can be used.

QuantumOptics.operators.DataOperator — Type.

Abstract type for operators with a data field.

This is an abstract type for operators that have a direct matrix representation stored in their .data field.

QuantumOptics.operators_dense.DenseOperator — Type.

DenseOperator(b1[, b2, data])

Dense array implementation of Operator.

The matrix consisting of complex floats is stored in the data field.

QuantumOptics.superoperators.DenseSuperOperator — Type.

DenseSuperOperator(b1[, b2, data])

SuperOperator stored as dense matrix.

QuantumOptics.fock.FockBasis — Type.

FockBasis(N)

Basis for a Fock space where N specifies a cutoff, i.e. what the highest included fock state is. Note that the dimension of this basis then is N+1.

QuantumOptics.bases.GenericBasis — Type.

GenericBasis(N)

A general purpose basis of dimension N.

Should only be used rarely since it defeats the purpose of checking that the bases of state vectors and operators are correct for algebraic operations. The preferred way is to specify special bases for different systems.

QuantumOptics.states.Ket — Type.

Ket(b::Basis[, data])

Ket state defined by coefficients in respect to the given basis.

QuantumOptics.operators_lazyproduct.LazyProduct — Type.

LazyProduct(operators[, factor=1])
LazyProduct(op1, op2...)

Lazy evaluation of products of operators.

The factors of the product are stored in the operators field. Additionally a complex factor is stored in the factor field which allows for fast multiplication with numbers.

QuantumOptics.operators_lazysum.LazySum — Type.

LazySum([factors,] operators)

Lazy evaluation of sums of operators.

All operators have to be given in respect to the same bases. The field factors accounts for an additional multiplicative factor for each operator stored in the operators field.

QuantumOptics.operators_lazytensor.LazyTensor — Type.

LazyTensor(b1[, b2], indices, operators[, factor=1])

Lazy implementation of a tensor product of operators.

The suboperators are stored in the operators field. The indices field specifies in which subsystem the corresponding operator lives. Additionally, a complex factor is stored in the factor field which allows for fast multiplication with numbers.

QuantumOptics.manybody.ManyBodyBasis — Type.

ManyBodyBasis(b, occupations)

Basis for a many body system.

The basis has to know the associated one-body basis b and which occupation states should be included. The occupations_hash is used to speed up checking if two many-body bases are equal.

QuantumOptics.particle.MomentumBasis — Type.

MomentumBasis(pmin, pmax, Npoints)
MomentumBasis(b::PositionBasis)

Basis for a particle in momentum space.

For simplicity periodic boundaries are assumed which means that pmax is not included in the basis but is defined to be the same as pmin.

When a PositionBasis is given as argument the exact values of $p_{min}$ and $p_{max}$ are due to the periodic boundary conditions more or less arbitrary and are chosen to be $-\pi/dx$ and $\pi/dx$ with $dx=(x_{max}-x_{min})/N$.

QuantumOptics.nlevel.NLevelBasis — Type.

NLevelBasis(N)

Basis for a system consisting of N states.

QuantumOptics.metrics.PPT — Function.

PPT(rho, index)

Peres-Horodecki criterion of partial transpose.

QuantumOptics.particle.PositionBasis — Type.

PositionBasis(xmin, xmax, Npoints)
PositionBasis(b::MomentumBasis)

Basis for a particle in real space.

For simplicity periodic boundaries are assumed which means that the rightmost point defined by xmax is not included in the basis but is defined to be the same as xmin.

When a MomentumBasis is given as argument the exact values of $x_{min}$ and $x_{max}$ are due to the periodic boundary conditions more or less arbitrary and are chosen to be $-\pi/dp$ and $\pi/dp$ with $dp=(p_{max}-p_{min})/N$.

QuantumOptics.QuantumOptics

QuantumOptics.operators_sparse.SparseOperator — Type.

SparseOperator(b1[, b2, data])

Sparse array implementation of Operator.

The matrix is stored as the julia built-in type SparseMatrixCSC in the data field.

QuantumOptics.superoperators.SparseSuperOperator — Type.

SparseSuperOperator(b1[, b2, data])

SuperOperator stored as sparse matrix.

QuantumOptics.spin.SpinBasis — Type.

SpinBasis(n)

Basis for spin-n particles.

The basis can be created for arbitrary spinnumbers by using a rational number, e.g. SpinBasis(3//2). The Pauli operators are defined for all possible spin numbers.

QuantumOptics.states.StateVector — Type.

Abstract base class for Bra and Ket states.

The state vector class stores the coefficients of an abstract state in respect to a certain basis. These coefficients are stored in the data field and the basis is defined in the basis field.

QuantumOptics.subspace.SubspaceBasis — Type.

SubspaceBasis(basisstates)

A basis describing a subspace embedded a higher dimensional Hilbert space.

QuantumOptics.superoperators.SuperOperator — Type.

Base class for all super operator classes.

Super operators are bijective mappings from operators given in one specific basis to operators, possibly given in respect to another, different basis. To embed super operators in an algebraic framework they are defined with a left hand basis basis_l and a right hand basis basis_r where each of them again consists of a left and right hand basis.

\[A_{bl_1,bl_2} = S_{(bl_1,bl_2) ↔ (br_1,br_2)} B_{br_1,br_2} \\ A_{br_1,br_2} = B_{bl_1,bl_2} S_{(bl_1,bl_2) ↔ (br_1,br_2)}\]

QuantumOptics.bases

QuantumOptics.bases.basis — Function.

basis(a)

Return the basis of an object.

If it's ambiguous, e.g. if an operator has a different left and right basis, an IncompatibleBases error is thrown.

QuantumOptics.states.basisstate — Function.

basisstate(b, index)

Basis vector specified by index as ket state.

For a composite system index can be a vector which then creates a tensor product state $|i_1⟩⊗|i_2⟩⊗…⊗|i_n⟩$ of the corresponding basis states.

basisstate(b::ManyBodyBasis, occupation::Vector{Int})

Return a ket state where the system is in the state specified by the given occupation numbers.

QuantumOptics.manybody.bosonstates — Function.

bosonstates(Nmodes, Nparticles)
bosonstates(b, Nparticles)

Generate all bosonic occupation states for N-particles in M-modes. Nparticles can be a vector to define a Hilbert space with variable particle number.

QuantumOptics.phasespace.coherentspinstate — Function.

coherentspinstate(b::SpinBasis, θ::Real, ϕ::Real)

A coherent spin state |θ,ϕ⟩ is analogous to the coherent state of the linear harmonic oscillator. Coherent spin states represent a collection of identical two-level systems and can be described by two angles θ and ϕ (although this parametrization is not unique), similarly to a qubit on the Bloch sphere.

QuantumOptics.fock.coherentstate — Function.

coherentstate(b::FockBasis, alpha)

Coherent state $|α⟩$ for the specified Fock space.

QuantumOptics.state_definitions.coherentthermalstate — Function.

coherentthermalstate(basis::FockBasis,H,T,alpha)

Coherent thermal state $D(α)exp(-H/T)/Tr[exp(-H/T)]D^†(α)$.

QuantumOptics.fock.create — Function.

create(b::FockBasis)

Creation operator for the specified Fock space.

create(b::ManyBodyBasis, index)

Creation operator for the i-th mode of the many-body basis b.

QuantumOptics.states.dagger — Function.

dagger(x)

Hermitian conjugate.

QuantumOptics.operators.dense — Function.

dense(op::AbstractOperator)

Convert an arbitrary Operator into a DenseOperator.

QuantumOptics.fock.destroy — Function.

destroy(b::FockBasis)

Annihilation operator for the specified Fock space.

destroy(b::ManyBodyBasis, index)

Annihilation operator for the i-th mode of the many-body basis b.

QuantumOptics.timeevolution.timeevolution_mcwf.diagonaljumps — Function.

diagonaljumps(rates, J)

Diagonalize jump operators.

The given matrix rates of decay rates is diagonalized and the corresponding set of jump operators is calculated.

Arguments

rates: Matrix of decay rates.

J: Vector of jump operators.

QuantumOptics.operators_sparse.diagonaloperator — Function.

diagonaloperator(b::Basis)

Create a diagonal operator of type SparseOperator.

QuantumOptics.fock.displace — Function.

displace(b::FockBasis, alpha)

Displacement operator $D(α)$ for the specified Fock space.

QuantumOptics.operators_dense.dm — Function.

dm(a::StateVector)

Create density matrix $|a⟩⟨a|$. Same as projector(a).

QuantumOptics.spectralanalysis.eigenenergies — Function.

eigenenergies(op::AbstractOperator[, n::Int; warning=true])

Calculate the lowest n eigenvalues.

This is just a thin wrapper around julia's eigvals. If more control about the way the calculation is done is needed, use the function directly. More details can be found at [http://docs.julialang.org/en/stable/stdlib/linalg/].

If the given operator is non-hermitian a warning is given. This behavior can be turned off using the keyword warning=false.

For sparse operators by default it only returns the 6 lowest eigenvalues.

QuantumOptics.spectralanalysis.eigenstates — Function.

eigenstates(op::AbstractOperator[, n::Int; warning=true])

Calculate the lowest n eigenvalues and their corresponding eigenstates.

This is just a thin wrapper around julia's eigen and eigs functions. Which of them is used depends on the type of the given operator. If more control about the way the calculation is done is needed, use the functions directly. More details can be found at [http://docs.julialang.org/en/stable/stdlib/linalg/].

NOTE: Especially for small systems full diagonalization with Julia's eigen function is often more desirable. You can convert a sparse operator A to a dense one using dense(A).

If the given operator is non-hermitian a warning is given. This behavior can be turned off using the keyword warning=false.

For sparse operators by default it only returns the 6 lowest eigenvalues.

QuantumOptics.operators.embed — Function.

embed(basis1[, basis2], indices::Vector, operators::Vector)

Tensor product of operators where missing indices are filled up with identity operators.

embed(basis1[, basis2], operators::Dict)

operators is a dictionary Dict{Vector{Int}, AbstractOperator}. The integer vector specifies in which subsystems the corresponding operator is defined.

QuantumOptics.metrics.entropy_vn — Function.

entropy_vn(rho)

Von Neumann entropy of a density matrix.

The Von Neumann entropy of a density operator is defined as

\[S(ρ) = -Tr(ρ \log(ρ)) = -\sum_n λ_n\log(λ_n)\]

where $λ_n$ are the eigenvalues of the density matrix $ρ$, $\log$ is the natural logarithm and $0\log(0) ≡ 0$.

Arguments

rho: Density operator of which to calculate Von Neumann entropy.

tol=1e-15: Tolerance for rounding errors in the computed eigenvalues.

QuantumOptics.eval

QuantumOptics.operators.expect — Function.

expect(op, state)

Expectation value of the given operator op for the specified state.

state can either be a (density) operator or a ket.

expect(index, op, state)

If an index is given, it assumes that op is defined in the subsystem specified by this number.

QuantumOptics.manybody.fermionstates — Function.

fermionstates(Nmodes, Nparticles)
fermionstates(b, Nparticles)

Generate all fermionic occupation states for N-particles in M-modes. Nparticles can be a vector to define a Hilbert space with variable particle number.

QuantumOptics.metrics.fidelity — Function.

fidelity(rho, sigma)

Fidelity of two density operators.

The fidelity of two density operators $ρ$ and $σ$ is defined by

\[F(ρ, σ) = Tr\left(\sqrt{\sqrt{ρ}σ\sqrt{ρ}}\right),\]

where $\sqrt{ρ}=\sum_n\sqrt{λ_n}|ψ⟩⟨ψ|$.

QuantumOptics.fock

QuantumOptics.fock.fockstate — Function.

fockstate(b::FockBasis, n)

Fock state $|n⟩$ for the specified Fock space.

QuantumOptics.particle.gaussianstate — Function.

gaussianstate(b::PositionBasis, x0, p0, sigma)
gaussianstate(b::MomentumBasis, x0, p0, sigma)

Create a Gaussian state around x0 andp0 with width sigma.

In real space the gaussian state is defined as

\[\Psi(x) = \frac{\sqrt{\Delta x}}{\pi^{1/4}\sqrt{\sigma}} e^{i p_0 (x-\frac{x_0}{2}) - \frac{(x-x_0)^2}{2 \sigma^2}}\]

and is connected to the momentum space definition

\[\Psi(p) = \frac{\sqrt{\sigma} \sqrt{\Delta x}}{\pi^{1/4}} e^{-i x_0 (p-\frac{p_0}{2}) - \frac{1}{2}(p-p_0)^2 \sigma^2}\]

via a Fourier-transformation

\[\Psi(p) = \frac{1}{\sqrt{2\pi}} \int_{-\infty}^{\infty} e^{-ipx}\Psi(x) \mathrm{d}x\]

The state has the properties

$⟨p⟩ = p_0$

$⟨x⟩ = x_0$

$\mathrm{Var}(x) = \frac{σ^2}{2}$

$\mathrm{Var}(p) = \frac{1}{2 σ^2}$

Due to the numerically necessary discretization additional scaling factora $\sqrt{Δx}$ and $\sqrt{Δp}$ are used so that $Ψx_i = \sqrt{Δ x} Ψ(x_i)$ and $Ψp_i = \sqrt{Δ x} Ψ(p_i)$ so that the resulting Ket state is normalized.

QuantumOptics.operators.identityoperator — Function.

identityoperator(a::Basis[, b::Basis])

Return an identityoperator in the given bases.

QuantumOptics.include

QuantumOptics.superoperators.liouvillian — Function.

liouvillian(H, J; rates, Jdagger)

Create a super-operator equivalent to the master equation so that $\dot ρ = S ρ$.

The super-operator $S$ is defined by

\[S ρ = -\frac{i}{ħ} [H, ρ] + \sum_i J_i ρ J_i^† - \frac{1}{2} J_i^† J_i ρ - \frac{1}{2} ρ J_i^† J_i\]

Arguments

H: Hamiltonian.

J: Vector containing the jump operators.

rates: Vector or matrix specifying the coefficients for the jump operators.

Jdagger: Vector containing the hermitian conjugates of the jump operators. If they are not given they are calculated automatically.

QuantumOptics.metrics.logarithmic_negativity — Function.

logarithmic_negativity(rho, index)

The logarithmic negativity of a density matrix ρ is defined as

\[N(ρ) = \log₂\|ρᵀ\|,\]

where ρᵀ is the partial transpose.

QuantumOptics.manybody

QuantumOptics.manybody.manybodyoperator — Function.

manybodyoperator(b::ManyBodyBasis, op)

Create the many-body operator from the given one-body operator op.

The given operator can either be a one-body operator or a two-body interaction. Higher order interactions are at the moment not implemented.

The mathematical formalism for the one-body case is described by

\[X = \sum_{ij} a_i^† a_j ⟨u_i| x | u_j⟩\]

and for the interaction case by

\[X = \sum_{ijkl} a_i^† a_j^† a_k a_l ⟨u_i|⟨u_j| x |u_k⟩|u_l⟩\]

where $X$ is the N-particle operator, $x$ is the one-body operator and $|u⟩$ are the one-body states associated to the different modes of the N-particle basis.

QuantumOptics.metrics

QuantumOptics.particle.momentum — Function.

momentum(b:MomentumBasis)

Momentum operator in momentum space.

momentum(b::PositionBasis)

Momentum operator in real space.

QuantumOptics.metrics.negativity — Function.

negativity(rho, index)

Negativity of rho with respect to subsystem index.

The negativity of a density matrix ρ is defined as

\[N(ρ) = \|ρᵀ\|,\]

where ρᵀ is the partial transpose.

QuantumOptics.nlevel

QuantumOptics.nlevel.nlevelstate — Function.

nlevelstate(b::NLevelBasis, n::Int)

State where the system is completely in the n-th level.

LinearAlgebra.norm — Function.

norm(x::StateVector)

Norm of the given bra or ket state.

LinearAlgebra.normalize — Function.

normalize(op)

Return the normalized operator so that its tr(op) is one.

normalize(x::StateVector)

Return the normalized state so that norm(x) is one.

LinearAlgebra.normalize! — Function.

normalize!(op)

In-place normalization of the given operator so that its tr(x) is one.

normalize!(x::StateVector)

In-place normalization of the given bra or ket so that norm(x) is one.

QuantumOptics.fock.number — Function.

number(b::FockBasis)

Number operator for the specified Fock space.

number(b::ManyBodyBasis, index)

Particle number operator for the i-th mode of the many-body basis b.

number(b::ManyBodyBasis)

Total particle number operator.

QuantumOptics.occupation

QuantumOptics.manybody.onebodyexpect — Function.

onebodyexpect(op, state)

Expectation value of the one-body operator op in respect to the many-body state.

QuantumOptics.operators

QuantumOptics.operators_dense

QuantumOptics.operators_lazyproduct

QuantumOptics.operators_lazysum

QuantumOptics.operators_lazytensor

QuantumOptics.operators_sparse

QuantumOptics.particle

QuantumOptics.state_definitions.passive_state — Function.

passive_state(rho,IncreasingEigenenergies::Bool=true)

Passive state $π$ of $ρ$. IncreasingEigenenergies=true means that higher indices correspond to higher energies.

QuantumOptics.bases.permutesystems — Function.

permutesystems(a, perm)

Change the ordering of the subsystems of the given object.

For a permutation vector [2,1,3] and a given object with basis [b1, b2, b3] this function results in [b2, b1, b3].

QuantumOptics.state_definitions.phase_average — Function.

phase_average(rho)

Returns the phase-average of $ρ$ containing only the diagonal elements.

QuantumOptics.phasespace

QuantumOptics.polynomials

Base.position — Function.

position(b::PositionBasis)

Position operator in real space.

position(b:MomentumBasis)

Position operator in momentum space.

QuantumOptics.particle.potentialoperator — Function.

potentialoperator(b::PositionBasis, V(x))

Operator representing a potential $V(x)$ in real space.

potentialoperator(b::MomentumBasis, V(x))

Operator representing a potential $V(x)$ in momentum space.

potentialoperator(b::CompositeBasis, V(x, y, z, ...))

Operator representing a potential $V$ in more than one dimension.

Arguments

b: Composite basis consisting purely either of PositionBasis or MomentumBasis. Note, that calling this with a composite basis in momentum space might consume a large amount of memory.

V: Function describing the potential. ATTENTION: The number of arguments accepted by V must match the spatial dimension. Furthermore, the order of the arguments has to match that of the order of the tensor product of bases (e.g. if b=bx⊗by⊗bz, then V(x,y,z)).

QuantumOptics.printing

QuantumOptics.operators_dense.projector — Function.

projector(b1, b2)

Projection operator between subspaces and superspaces or between two subspaces.

projector(a::Ket, b::Bra)

Projection operator $|a⟩⟨b|$.

projector(a::Ket)

Projection operator $|a⟩⟨a|$.

projector(a::Bra)

Projection operator $|a⟩⟨a|$.

QuantumOptics.bases.ptrace — Function.

ptrace(a, indices)

Partial trace of the given basis, state or operator.

The indices argument, which can be a single integer or a vector of integers, specifies which subsystems are traced out. The number of indices has to be smaller than the number of subsystems, i.e. it is not allowed to perform a full trace.

QuantumOptics.metrics.ptranspose — Function.

ptranspose(rho, index)

Partial transpose of rho with respect to subspace specified by index.

QuantumOptics.phasespace.qfunc — Function.

qfunc(a, α)
qfunc(a, x, y)
qfunc(a, xvec, yvec)

Husimi Q representation $⟨α|ρ|α⟩/π$ for the given state or operator a. The function can either be evaluated on one point α or on a grid specified by the vectors xvec and yvec. Note that conversion from x and y to α is done via the relation $α = \frac{1}{\sqrt{2}}(x + i y)$.

QuantumOptics.phasespace.qfuncsu2 — Function.

qfuncsu2(ket,Ntheta;Nphi=2Ntheta)
qfuncsu2(rho,Ntheta;Nphi=2Ntheta)

Husimi Q SU(2) representation $⟨θ,ϕ|ρ|θ,ϕ⟩/π$ for the given state.

The function calculates the SU(2) Husimi representation of a state on the generalised bloch sphere (0 < θ < π and 0 < ϕ < 2 π) with a given resolution (Ntheta, Nphi).

qfuncsu2(rho,θ,ϕ)
qfuncsu2(ket,θ,ϕ)

This version calculates the Husimi Q SU(2) function at a position given by θ and ϕ.

QuantumOptics.state_definitions.randoperator — Function.

randoperator(b1[, b2])

Calculate a random unnormalized dense operator.

QuantumOptics.state_definitions.randstate — Function.

randstate(basis)

Calculate a random normalized ket state.

QuantumOptics.particle.samplepoints — Function.

samplepoints(b::PositionBasis)

x values of the real space basis.

samplepoints(b::MomentumBasis)

p values of the momentum basis.

QuantumOptics.semiclassical

QuantumOptics.spin.sigmam — Function.

sigmam(b::SpinBasis)

Lowering operator $σ_-$ for the given Spin basis.

QuantumOptics.spin.sigmap — Function.

sigmap(b::SpinBasis)

Raising operator $σ_+$ for the given Spin basis.

QuantumOptics.spin.sigmax — Function.

sigmax(b::SpinBasis)

Pauli $σ_x$ operator for the given Spin basis.

QuantumOptics.spin.sigmay — Function.

sigmay(b::SpinBasis)

Pauli $σ_y$ operator for the given Spin basis.

QuantumOptics.spin.sigmaz — Function.

sigmaz(b::SpinBasis)

Pauli $σ_z$ operator for the given Spin basis.

QuantumOptics.spectralanalysis.simdiag — Function.

simdiag(ops; atol, rtol)

Simultaneously diagonalize commuting Hermitian operators specified in ops.

This is done by diagonalizing the sum of the operators. The eigenvalues are computed by $a = ⟨ψ|A|ψ⟩$ and it is checked whether the eigenvectors fulfill the equation $A|ψ⟩ = a|ψ⟩$.

Arguments

ops: Vector of sparse or dense operators.

atol=1e-14: kwarg of Base.isapprox specifying the tolerance of the approximate check

rtol=1e-14: kwarg of Base.isapprox specifying the tolerance of the approximate check

Returns

evals_sorted: Vector containing all vectors of the eigenvalues sorted by the eigenvalues of the first operator.

v: Common eigenvectors.

QuantumOptics.sortedindices

SparseArrays.sparse — Function.

sparse(op::AbstractOperator)

Convert an arbitrary operator into a SparseOperator.

QuantumOptics.sparsematrix

QuantumOptics.spectralanalysis

QuantumOptics.spin

QuantumOptics.spin.spindown — Function.

spindown(b::SpinBasis)

Spin down state for the given Spin basis.

QuantumOptics.spin.spinup — Function.

spinup(b::SpinBasis)

Spin up state for the given Spin basis.

QuantumOptics.superoperators.spost — Function.

Create a super-operator equivalent for left side operator multiplication.

For operators $A$, $B$ the relation

\[ \mathrm{spost}(A) B = B A\]

holds. op can be a dense or a sparse operator.

QuantumOptics.superoperators.spre — Function.

spre(op)

Create a super-operator equivalent for right side operator multiplication.

For operators $A$, $B$ the relation

\[ \mathrm{spre}(A) B = A B\]

holds. op can be a dense or a sparse operator.

QuantumOptics.state_definitions

QuantumOptics.states

QuantumOptics.steadystate

QuantumOptics.stochastic

QuantumOptics.subspace

QuantumOptics.superoperators

QuantumOptics.bases.tensor — Function.

tensor(x::AbstractOperator, y::AbstractOperator, z::AbstractOperator...)

Tensor product $\hat{x}⊗\hat{y}⊗\hat{z}⊗…$ of the given operators.

tensor(x::Ket, y::Ket, z::Ket...)

Tensor product $|x⟩⊗|y⟩⊗|z⟩⊗…$ of the given states.

tensor(x, y, z...)

Tensor product of the given objects. Alternatively, the unicode symbol ⊗ (\otimes) can be used.

tensor(x::Basis, y::Basis, z::Basis...)

Create a CompositeBasis from the given bases.

Any given CompositeBasis is expanded so that the resulting CompositeBasis never contains another CompositeBasis.

tensor(x::Ket, y::Bra)

Outer product $|x⟩⟨y|$ of the given states.

QuantumOptics.state_definitions.thermalstate — Function.

thermalstate(H,T)

Thermal state $exp(-H/T)/Tr[exp(-H/T)]$.

QuantumOptics.timecorrelations

QuantumOptics.timeevolution

LinearAlgebra.tr — Function.

tr(x::AbstractOperator)

Trace of the given operator.

QuantumOptics.metrics.tracedistance — Function.

tracedistance(rho, sigma)

Trace distance between rho and sigma.

It is defined as

\[T(ρ,σ) = \frac{1}{2} Tr\{\sqrt{(ρ - σ)^† (ρ - σ)}\}.\]

It calls tracenorm which in turn either uses tracenorm_h or tracenorm_nh depending if $ρ-σ$ is hermitian or not.

QuantumOptics.metrics.tracedistance_h — Function.

tracedistance_h(rho, sigma)

Trace distance between rho and sigma.

It uses the identity

\[T(ρ,σ) = \frac{1}{2} Tr\{\sqrt{(ρ - σ)^† (ρ - σ)}\} = \frac{1}{2} \sum_i |λ_i|\]

where $λ_i$ are the eigenvalues of rho - sigma.

QuantumOptics.metrics.tracedistance_nh — Function.

tracedistance_nh(rho, sigma)

Trace distance between rho and sigma.

Note that in this case rho and sigma don't have to be represented by square matrices (i.e. they can have different left-hand and right-hand bases).

It uses the identity

\[ T(ρ,σ) = \frac{1}{2} Tr\{\sqrt{(ρ - σ)^† (ρ - σ)}\} = \frac{1}{2} \sum_i σ_i\]

where $σ_i$ are the singular values of rho - sigma.

QuantumOptics.metrics.tracenorm — Function.

tracenorm(rho)

Trace norm of rho.

It is defined as

\[T(ρ) = Tr\{\sqrt{ρ^† ρ}\}.\]

Depending if rho is hermitian either tracenorm_h or tracenorm_nh is called.

QuantumOptics.metrics.tracenorm_h — Function.

tracenorm_h(rho)

Trace norm of rho.

It uses the identity

\[T(ρ) = Tr\{\sqrt{ρ^† ρ}\} = \sum_i |λ_i|\]

where $λ_i$ are the eigenvalues of rho.

QuantumOptics.metrics.tracenorm_nh — Function.

tracenorm_nh(rho)

Trace norm of rho.

Note that in this case rho doesn't have to be represented by a square matrix (i.e. it can have different left-hand and right-hand bases).

It uses the identity

\[ T(ρ) = Tr\{\sqrt{ρ^† ρ}\} = \sum_i σ_i\]

where $σ_i$ are the singular values of rho.

QuantumOptics.particle.transform — Function.

transform(b1::PositionBasis, b2::FockBasis; x0=1)
transform(b1::FockBasis, b2::PositionBasis; x0=1)

Transformation operator between position basis and fock basis.

The coefficients are connected via the relation

\[ψ(x_i) = \sum_{n=0}^N ⟨x_i|n⟩ ψ_n\]

where $⟨x_i|n⟩$ is the value of the n-th eigenstate of a particle in a harmonic trap potential at position $x$, i.e.:

\[⟨x_i|n⟩ = π^{-\frac{1}{4}} \frac{e^{-\frac{1}{2}\left(\frac{x}{x_0}\right)^2}}{\sqrt{x_0}} \frac{1}{\sqrt{2^n n!}} H_n\left(\frac{x}{x_0}\right)\]

transform(b1::MomentumBasis, b2::PositionBasis)
transform(b1::PositionBasis, b2::MomentumBasis)

Transformation operator between position basis and momentum basis.

transform(b1::CompositeBasis, b2::CompositeBasis)

Transformation operator between two composite bases. Each of the bases has to contain bases of type PositionBasis and the other one a corresponding MomentumBasis.

QuantumOptics.transformations

QuantumOptics.nlevel.transition — Function.

transition(b::ManyBodyBasis, to::Int, from::Int)

Operator $|\mathrm{to}⟩⟨\mathrm{from}|$ transferring particles between modes.

transition(b::NLevelBasis, to::Int, from::Int)

Transition operator $|\mathrm{to}⟩⟨\mathrm{from}|$.

QuantumOptics.operators.variance — Function.

variance(op, state)

Variance of the given operator op for the specified state.

state can either be a (density) operator or a ket.

variance(index, op, state)

If an index is given, it assumes that op is defined in the subsystem specified by this number

QuantumOptics.phasespace.wigner — Function.

wigner(a, α)
wigner(a, x, y)
wigner(a, xvec, yvec)

Wigner function for the given state or operator a. The function can either be evaluated on one point α or on a grid specified by the vectors xvec and yvec. Note that conversion from x and y to α is done via the relation $α = \frac{1}{\sqrt{2}}(x + i y)$.

QuantumOptics.phasespace.wignersu2 — Function.

wignersu2(ket,Ntheta;Nphi=2Ntheta)
wignersu2(rho,Ntheta;Nphi=2Ntheta)

Wigner SU(2) representation for the given state with a resolution (Ntheta, Nphi).

The function calculates the SU(2) Wigner representation of a state on the generalised bloch sphere (0 < θ < π and 0 < ϕ < 2 π) with a given resolution by decomposing the state into the basis of spherical harmonics.

wignersu2(rho,θ,ϕ)
wignersu2(ket,θ,ϕ)

This version calculates the Wigner SU(2) function at a position given by θ and ϕ

QuantumOptics.phasespace.ylm — Function.

ylm(l::Integer,m::Integer,theta::Real,phi::Real)

Spherical harmonics Y(l,m)(θ,ϕ) where l ∈ N, m = -l,-l+1,...,l-1,l, θ ∈ [0,π], and ϕ ∈ [0,2π).

This function calculates the value of Y(l,m) spherical harmonic at position θ and ϕ.

QuantumOptics.⊗