Docstrings

CodeTools.CodeTools

CodeTools.MODULE_STARTERS

CodeTools.SCOPE_STARTERS

CodeTools._names

CodeTools._stdlibs

CodeTools.accessible

CodeTools.allchildren

CodeTools.builtin_completions

CodeTools.builtins

CodeTools.children

CodeTools.codemodule — Function.

Takes Julia source code and a line number, gives back the string name of the module at that line.

CodeTools.completions

CodeTools.completiontype

CodeTools.description

CodeTools.dirs

CodeTools.dirsnearby — Function.

Takes a start directory and returns a set of nearby directories.

CodeTools.doc

CodeTools.eval

CodeTools.filemodule

CodeTools.filemodule_ — Function.

Takes an absolute path to a file and returns a string representing the module it belongs to.

CodeTools.files

CodeTools.filtervalid

CodeTools.find_include — Function.

Takes an absolute path to a file and returns the (file, line) where that file is included or nothing.

CodeTools.flat_content

CodeTools.flat_content!

CodeTools.flatten

CodeTools.fullsignature

CodeTools.funcprefix

CodeTools.getmodule

CodeTools.getpackage

CodeTools.getsubmod

CodeTools.getthing

CodeTools.hasdoc

CodeTools.identifier_pattern

CodeTools.include

CodeTools.includeline — Function.

Takes a given Julia source file and another (absolute) path, gives the line on which the path is included in the file or 0.

CodeTools.includepath

CodeTools.isdir′

CodeTools.isfile′

CodeTools.jl_files

CodeTools.lines

CodeTools.memoize

CodeTools.memoize_debounce

CodeTools.meta_cache

CodeTools.methodsorwith

CodeTools.moduleusings

CodeTools.namecompletions

CodeTools.namecompletions_

CodeTools.pathcompletions

CodeTools.pathmeta

CodeTools.pathprefix

CodeTools.pkgcompletions

CodeTools.pkgmeta

CodeTools.prefix

CodeTools.prefix_pattern

CodeTools.providers

CodeTools.readdir′

CodeTools.returns

CodeTools.signature

CodeTools.stdlibs

CodeTools.stringmeta

CodeTools.texcommands

CodeTools.thingorfunc

CodeTools.trygetdoc

CodeTools.withmeta

CodeTools.withpath