Docstrings

SASLib.CIDict

SASLib.Column

SASLib.ColumnNamePointer

SASLib.ConfigError

SASLib.EMPTY_STRING

SASLib.ENCODINGS_OK_WITH_BASE_TRANSCODER

SASLib.FALLBACK_ENCODING

SASLib.FileFormatError

SASLib.Handler

SASLib.Metadata — Type.

Metadata contains information about a SAS data file.

Fields

filename: file name/path of the SAS data set

encoding: file encoding e.g. "ISO8859-1"

endianness: either :LittleEndianor:BigEndian`

compression: could be :RLE, :RDC, or :none

pagesize: size of each data page in bytes

npages: number of pages in the file

nrows: number of data rows in the file

ncols: number of data columns in the file

columnsinfo: vector of column symbols and their respective types

SASLib.ObjectPool — Type.

ObjectPool is a fixed-size one-dimensional array that does not store any duplicate copies of the same object. So the benefit is space-efficiency. The tradeoff is the time used to maintain the index. This is useful for denormalized data frames where string values may be repeated many times.

An ObjectPool must be initialize with a default value and a fixed array size. If your requirement does not fit such assumptions, you may want to look into using PooledArrays or CategoricalArrays package instead.

The implementation is very primitive and is tailor for application that knows exactly how much memory to allocate.

SASLib.REGULAR_STR_ARRAY

SASLib.ReaderConfig

SASLib.ResultSet — Type.

ResultSet is the primary object that represents data returned from reading a SAS data file. ResultSet implements the Base.Iteration interface as well as the IterableTables.jl interface.

Fields

columns: a vector of columns, each being a vector itself

names: a vector of column symbols

size: a tuple (nrows, ncols)

Accessors

columns(::ResultSet)

names(::ResultSet)

size(::ResultSet)

size(::ResultSet, dim::Integer)

Single Row/Column Indexing

rs[i] returns a tuple for row i

rs[:c] returns a vector for column c

Multiple Row/Column Indexing

rs[i:j] returns a view of ResultSet with rows between i and j

rs[:c...] returns a view of ResultSet with columns specified e.g. rs[:A, :B]

Cell Indexing

rs[i,j] returns a single value for row i column j

rs[i,:c] returns a single value for row i column c

Specific cell can be assigned using the above indexing methods

SASLib.SASLib

SASLib.SAS_RLE_COMMAND_COPY1

SASLib.SAS_RLE_COMMAND_COPY17

SASLib.SAS_RLE_COMMAND_COPY33

SASLib.SAS_RLE_COMMAND_COPY49

SASLib.SAS_RLE_COMMAND_COPY64

SASLib.SAS_RLE_COMMAND_INSERT_AT17

SASLib.SAS_RLE_COMMAND_INSERT_AT2

SASLib.SAS_RLE_COMMAND_INSERT_BLANK17

SASLib.SAS_RLE_COMMAND_INSERT_BLANK2

SASLib.SAS_RLE_COMMAND_INSERT_BYTE18

SASLib.SAS_RLE_COMMAND_INSERT_BYTE3

SASLib.SAS_RLE_COMMAND_INSERT_ZERO17

SASLib.SAS_RLE_COMMAND_INSERT_ZERO2

SASLib.SubHeaderPointer

SASLib.VENDOR_SAS

SASLib.VENDOR_STAT_TRANSFER

SASLib._chunk_to_dataframe

SASLib._determine_vendor

SASLib._get_subheader_index

SASLib._metadata

SASLib._open

SASLib._process_columnattributes_subheader

SASLib._process_columnlist_subheader

SASLib._process_columnname_subheader

SASLib._process_columnsize_subheader

SASLib._process_columntext_subheader

SASLib._process_format_subheader

SASLib._process_page_meta

SASLib._process_page_metadata

SASLib._process_rowsize_subheader

SASLib._process_subheader

SASLib._process_subheader_counts

SASLib._process_subheader_pointers

SASLib._read_bytes

SASLib._read_float

SASLib._read_int

SASLib._read_next_page_content

SASLib._read_page_header

SASLib._read_subheader_signature

SASLib.align_1_checker_value

SASLib.align_1_length

SASLib.align_1_offset

SASLib.align_1_value

SASLib.align_2_length

SASLib.align_2_offset

SASLib.align_2_value

SASLib.block_count_length

SASLib.block_count_offset

SASLib.brstrip — Function.

Strip from the right end of the bytes array for any byte that matches the ones specified in the remove argument. See Python's bytes.rstrip function.

SASLib.case_insensitive_in

SASLib.check_user_column_types

SASLib.close — Function.

close(handler::Handler)

Close the handler object. This function effectively closes the underlying iostream. It must be called after the program finished reading data.

This function is needed only when SASLib.open and SASLib.read functions are used instead of the more convenient readsas function.

SASLib.col_count_p1_multiplier

SASLib.col_count_p2_multiplier

SASLib.colfmt

SASLib.column_data_length_length

SASLib.column_data_length_offset

SASLib.column_data_offset_offset

SASLib.column_name_length_length

SASLib.column_name_length_offset

SASLib.column_name_offset_length

SASLib.column_name_offset_offset

SASLib.column_name_pointer_length

SASLib.column_name_text_subheader_length

SASLib.column_name_text_subheader_offset

SASLib.column_type_decimal

SASLib.column_type_length

SASLib.column_type_none

SASLib.column_type_offset

SASLib.column_type_string

SASLib.columns

SASLib.compression_method_none

SASLib.compression_method_rdc

SASLib.compression_method_rle

SASLib.contains — Function.

Find needle in the haystack with both Vector{UInt8} type arguments.

SASLib.convert_column_type_if_needed!

SASLib.convertfloat64f!

SASLib.convertint64B

SASLib.convertint64L

SASLib.createnumarray

SASLib.createstrarray

SASLib.current_page_type_str

SASLib.dataset_length

SASLib.dataset_offset

SASLib.date_created_length

SASLib.date_created_offset

SASLib.date_from_float

SASLib.date_modified_length

SASLib.date_modified_offset

SASLib.datetime_from_float

SASLib.decode_string

SASLib.default_chunk_size

SASLib.default_verbose_level

SASLib.encoding_length

SASLib.encoding_names

SASLib.encoding_offset

SASLib.endianness_length

SASLib.endianness_offset

SASLib.eval

SASLib.file_type_length

SASLib.file_type_offset

SASLib.header_size_length

SASLib.header_size_offset

SASLib.include

SASLib.index_columnAttributesIndex

SASLib.index_columnListIndex

SASLib.index_columnNameIndex

SASLib.index_columnSizeIndex

SASLib.index_columnTextIndex

SASLib.index_dataSubheaderIndex

SASLib.index_end_of_header

SASLib.index_formatAndLabelIndex

SASLib.index_rowSizeIndex

SASLib.index_subheaderCountsIndex

SASLib.init_handler

SASLib.issym

SASLib.lastcharpos

SASLib.lcase

SASLib.logdebug

SASLib.magic

SASLib.metadata — Function.

Return metadata of a SAS data file. See SASLib.Metadata.

SASLib.my_read_next_page

SASLib.open — Function.

open(filename::AbstractString; encoding::AbstractString = "", convertdates::Bool = true, includecolumns::Vector = [], excludecolumns::Vector = [], stringarrayfn::Dict = Dict(), numberarrayfn::Dict = Dict(), columntypes::Dict = Dict{Symbol,Type}(), verbose_level::Int64 = 1)

Open a SAS7BDAT data file. Returns a SASLib.Handler object that can be used in the subsequent SASLib.read and SASLib.close functions.

SASLib.os_maker_length

SASLib.os_maker_offset

SASLib.os_name_length

SASLib.os_name_offset

SASLib.os_version_number_length

SASLib.os_version_number_offset

SASLib.page_amd_type

SASLib.page_bit_offset_x64

SASLib.page_bit_offset_x86

SASLib.page_comp_type

SASLib.page_count_length

SASLib.page_count_offset

SASLib.page_data_type

SASLib.page_meta_data_mix_types

SASLib.page_meta_type

SASLib.page_metc_type

SASLib.page_mix_types

SASLib.page_size_length

SASLib.page_size_offset

SASLib.page_type_length

SASLib.page_type_offset

SASLib.page_type_str

SASLib.platform_length

SASLib.platform_offset

SASLib.populate_column_indices

SASLib.populate_column_names

SASLib.println1

SASLib.println2

SASLib.println3

SASLib.process_byte_array_with_data

SASLib.rdc_compression

SASLib.rdc_decompress

SASLib.read — Function.

read(handler::Handler, nrows=0)

Read data from the handler (see SASLib.open). If nrows is not specified, read the entire file content. When called again, fetch the next nrows rows.

SASLib.read_chunk

SASLib.read_data

SASLib.read_file_metadata

SASLib.read_first_page

SASLib.read_header

SASLib.read_next_page

SASLib.readline

SASLib.readsas — Function.

readsas(filename::AbstractString; encoding::AbstractString = "", convertdates::Bool = true, includecolumns::Vector = [], excludecolumns::Vector = [], stringarrayfn::Dict = Dict(), numberarrayfn::Dict = Dict(), columntypes::Dict = Dict{Symbol,Type}(), verbose_level::Int64 = 1)

Read a SAS7BDAT file.

Encoding may be used as an override only if the file cannot be read using the encoding specified in the file. If you receive a warning about unknown encoding then check your system's supported encodings from the iconv library e.g. using the iconv --list command.

If convert_dates == false then no conversion is made and you will get the number of days for Date columns (or number of seconds for DateTime columns) since 1-JAN-1960.

By default, all columns will be read. If you only need a subset of the columns, you may specify either include_columns or exclude_columns but not both. They are just arrays of columns indices or symbols e.g. [1, 2, 3] or [:employeeid, :firstname, :lastname]

String columns by default are stored in SASLib.ObjectPool, which is an array-like structure that is more space-efficient when there is a high number of duplicate values. However, if there are too many unique items (> 10%) then it's automatically switched over to a regular Array.

If you wish to use a different kind of array, you can pass your array constructor via the string_array_fn dict. The constructor must take a single integer argument that represents the size of the array. The convenient REGULAR_STR_ARRAY function can be used if you just want to use the regular Array{String} type.

For examples, string_array_fn = Dict(:column1 => (n)->CategoricalArray{String}((n,))) or string_array_fn = Dict(:column1 => REGULAR_STR_ARRAY).

For numeric columns, you may specify your own array constructors using the number_array_fn parameter. Perhaps you have a different kind of array to store the values e.g. SharedArray.

Specify column_type argument if any conversion is required. It should be a Dict, mapping column symbol to a data type.

For debugging purpose, verbose_level may be set to a value higher than 1. Verbose level 0 will output nothing to the console, essentially a total quiet option.

SASLib.rle_compression

SASLib.rle_decompress

SASLib.row_count_offset_multiplier

SASLib.row_count_on_mix_page_offset_multiplier

SASLib.row_length_offset_multiplier

SASLib.sas_date_formats

SASLib.sas_date_origin

SASLib.sas_datetime_formats

SASLib.sas_datetime_origin

SASLib.sas_release_length

SASLib.sas_release_offset

SASLib.sas_server_type_length

SASLib.sas_server_type_offset

SASLib.seven_bit_data

SASLib.sizestr

SASLib.subheader_comp_compressed

SASLib.subheader_comp_truncated

SASLib.subheader_comp_uncompressed

SASLib.subheader_count_length

SASLib.subheader_count_offset

SASLib.subheader_pointer_length_x64

SASLib.subheader_pointer_length_x86

SASLib.subheader_pointers_offset

SASLib.subheader_signature_to_index

SASLib.symindex

SASLib.text_block_size_length

SASLib.transcode_data

SASLib.transcode_metadata

SASLib.typesfmt

SASLib.typesof

SASLib.u64_byte_checker_value

SASLib.zero_space