Docstrings

Weave.AsciiDoc

Weave.ChunkOutput

Weave.CodeChunk

Weave.CollectResult

Weave.DocChunk

Weave.GithubMarkdown

Weave.Hugo — Type.

Formatter for Hugo: https://gohugo.io/

When uglyURLs is false, prepend figure path by ...

Weave.Inline

Weave.InlineCode

Weave.InlineText

Weave.JMarkdown2HTML

Weave.JMarkdown2tex

Weave.Markdown2HTML

Weave.MarkdownOutput

Weave.MarkupInput

Weave.MultiMarkdown

Weave.NotebookInput

Weave.NotebookOutput

Weave.NowebOutput

Weave.Pandoc

Weave.Pandoc2HTML

Weave.Report

Weave.Rest

Weave.ScriptInput

Weave.ScriptOutput

Weave.ScriptResult

Weave.TermResult

Weave.Tex

Weave.Weave

Weave.WeaveChunk

Weave.WeaveDoc

Weave.__init__

Weave.add_figure — Function.

Add saved figure name to results and return the name

Weave.adoc

Weave.capture_output

Weave.clear_sandbox

Weave.collect_results

Weave.convert_doc — Function.

convert_doc(infile::AbstractString, outfile::AbstractString; format = nothing)

Convert Weave documents between different formats

infile = Name of the input document

outfile = Name of the output document

format = Output format (optional). Detected from outfile extension, but can be set to "script", "markdown", "notebook" or "noweb".

Convert Weave document to Jupyter notebook format

Convert Weave document to Jupyter notebook format

Convert Weave document to noweb format

Convert Weave document to script format

Weave.defaultParams

Weave.default_mime_types

Weave.detect_doctype — Function.

Detect the output format based on file extension

Weave.detect_informat — Function.

Detect the input format based on file extension

Weave.detect_outformat — Function.

Autodetect format for converter

Weave.docParams

Weave.embed_figures

Weave.eval

Weave.eval_chunk

Weave.format

Weave.format_chunk

Weave.format_code

Weave.format_inline

Weave.format_output

Weave.format_termchunk

Weave.formatfigures

Weave.formats

Weave.get_chunk_defaults — Function.

get_chunk_defaults()

Get default options used for code chunks.

Weave.get_cwd

Weave.get_figname

Weave.get_outname — Function.

Get output file name based on out_path

Get output file name based on out_path

Weave.getstdout

Weave.github

Weave.highlight

Weave.hugo

Weave.img2base64

Weave.include

Weave.include_weave — Function.

include_weave(doc, informat=:auto)

Include code from Weave document calling include_string on all code from doc. Code is run in the path of the include document.

Weave.indent

Weave.input_formats

Weave.list_out_formats — Function.

list_out_formats()

List supported output formats

Weave.md2html

Weave.md2tex

Weave.multimarkdown

Weave.notebook — Function.

notebook(source::String, out_path=:pwd)

Convert Weave document source to Jupyter notebook and execute the code using nbconvert. Requires IJulia. Ignores all chunk options

out_path: Path where the output is generated. Can be: :doc: Path of the source document, :pwd: Julia working directory, "somepath": Path as a String e.g "/home/mpastell/weaveout"

nbconvert cell timeout in seconds. Defaults to -1 (no timeout)

Weave.output_formats

Weave.pandoc

Weave.pandoc2html — Function.

pandoc2html(formatted::AbstractString, doc::WeaveDoc)

Convert output from pandoc markdown to html using Weave.jl template

Weave.pandoc2pdf — Function.

pandoc2pdf(formatted::AbstractString, doc::WeaveDoc)

Convert output from pandoc markdown to pdf using Weave.jl template

Weave.parse_doc — Function.

Parse documents with Weave.jl markup

Parse .jl scripts with Weave.jl markup

Parse IJUlia notebook

Weave.parse_header

Weave.parse_inline

Weave.parse_input

Weave.pdoc2html

Weave.pop_postexecute_hook

Weave.pop_preexecute_hook

Weave.postexecute_hooks

Weave.preexecute_hooks

Weave.push_postexecute_hook

Weave.push_preexecute_hook

Weave.pushopt

Weave.rcParams

Weave.read_cache

Weave.read_doc — Function.

Read and parse input document

Weave.render_doc — Function.

render_doc(formatted::AbstractString, format)

Render formatted document to a template

Weave.reset_report

Weave.restore_chunk

Weave.restore_chunk_defaults — Function.

restore_chunk_defaults()

Restore Weave.jl default chunk options

Weave.rst

Weave.run_chunk

Weave.run_code

Weave.run_inline

Weave.run_latex

Weave.set_chunk_defaults — Function.

set_chunk_defaults(opts::Dict{Symbol, Any})

Set default options for code chunks, use getchunkdefaults to see the current values.

e.g. set default dpi to 200 and fig_width to 8

julia> set_chunk_defaults(Dict{Symbol, Any}(:dpi => 200, fig_width => 8))

Weave.set_rc_params

Weave.strip_header

Weave.stylesheet

Weave.tangle — Function.

tangle(source ; out_path=:doc, informat="noweb")

Tangle source code from input document to .jl file.

informat: "noweb" of "markdown"

out_path: Path where the output is generated. Can be: :doc: Path of the source document, :pwd: Julia working directory, "somepath", directory name as a string e.g "/home/mpastell/weaveout"

or filename as string e.g. ~/outpath/outfile.jl.

Weave.tex

Weave.texminted

Weave.uc2tex

Weave.weave — Function.

weave(source ; doctype = :auto,
    informat=:auto, out_path=:doc, args = Dict(),
    mod::Union{Module, Symbol} = Main,
    fig_path = "figures", fig_ext = nothing,
    cache_path = "cache", cache=:off,
    template = nothing, highlight_theme = nothing, css = nothing,
    pandoc_options = "",
    latex_cmd = "xelatex")

Weave an input document to output file.

doctype: :auto = set based on file extension or specify one of the supported formats. See list_out_formats()

informat: :auto = set based on file extension or set to "noweb", "markdown" or script

out_path: Path where the output is generated. Can be: :doc: Path of the source document, :pwd: Julia working directory, "somepath": output directory as a String e.g "/home/mpastell/weaveout" or filename as string e.g. ~/outpath/outfile.tex.

args: dictionary of arguments to pass to document. Available as WEAVE_ARGS

mod: Module where Weave evals code. Defaults to :sandbox to create new sandbox module, you can also pass a module e.g. Main.

fig_path: where figures will be generated, relative to out_path

fig_ext: Extension for saved figures e.g. ".pdf", ".png". Default setting depends on doctype.

cache_path: where of cached output will be saved.

cache: controls caching of code: :off = no caching, :all = cache everything, :user = cache based on chunk options, :refresh, run all code chunks and save new cache.

throw_errors if false errors are included in output document and the whole document is executed. if true errors are thrown when they occur.

template : Template (file path) for md2html or md2tex formats.

highlight_theme : Theme (Highlights.AbstractTheme) for used syntax highlighting

css : CSS (file path) used for md2html format

pandoc_options = String array of options to pass to pandoc for pandoc2html and pandoc2pdf formats e.g. ["–toc", "-N"]

latex_cmd the command used to make pdf from .tex

Note: Run Weave from terminal and not using IJulia, Juno or ESS, they tend to mess with capturing output.

Weave.wrapline

Weave.wraplines

Weave.write_cache