Docstrings

TensorOperations.@dividebody

TensorOperations.@optimalcontractiontree

TensorOperations.@stridedloops

TensorOperations.@tensor — Macro.

@tensor(block)

Specify one or more tensor operations using Einstein's index notation. Indices can be chosen to be arbitrary Julia variable names, or integers. When contracting several tensors together, this will be evaluated as pairwise contractions in left to right order, unless the so-called NCON style is used (positive integers for contracted indices and negative indices for open indices).

TensorOperations.@tensoropt — Macro.

@tensoropt(optex, block)
@tensoropt(block)

Specify one or more tensor operations using Einstein's index notation. Indices can be chosen to be arbitrary Julia variable names, or integers. When contracting several tensors together, the macro will determine (at compile time) the optimal contraction order depending on the cost associated to the individual indices. If no optex is provided, all indices are assumed to have an abstract scaling χ which is optimized in the asympotic limit of large χ.

The cost can be specified in the following ways:

@tensoropt (a=>χ,b=>χ^2,c=>2*χ,e=>5) C[a,b,c,d] := A[a,e,c,f,h]*B[f,g,e,b]*C[g,d,h]
# asymptotic cost as specified for listed indices, unlisted indices have cost 1 (any symbol for χ can be used)
@tensoropt (a,b,c,e) C[a,b,c,d] := A[a,e,c,f,h]*B[f,g,e,b]*C[g,d,h]
# asymptotic cost χ for indices a,b,c,e, other indices (d,f) have cost 1
@tensoropt !(a,b,c,e) C[a,b,c,d] := A[a,e,c,f,h]*B[f,g,e,b]*C[g,d,h]
# cost 1 for indices a,b,c,e; other indices (d,f) have asymptotic cost χ
@tensoropt C[a,b,c,d] := A[a,e,c,f,h]*B[f,g,e,b]*C[g,d,h]
# asymptotic cost χ for all indices (a,b,c,d,e,f)

Note that @tensoropt will optimize any tensor contraction sequence it encounters in the (block of) expressions. It will however not break apart expressions that have been explicitly grouped with parenthesis, i.e. in

@tensoroptC[a,b,c,d] := A[a,e,c,f,h]*(B[f,g,e,b]*C[g,d,h])

it will always contract B and C first. For a single tensor contraction sequence, the optimal contraction order and associated (asymptotic) cost can be obtained using @optimalcontractiontree.

TensorOperations.AbstractPoly

TensorOperations.BASELENGTH

TensorOperations.ConjugatedStridedData

TensorOperations.IndexError

TensorOperations.IndexTuple

TensorOperations.NormalStridedData

TensorOperations.One

TensorOperations.Poly

TensorOperations.Power

TensorOperations.StaticLength

TensorOperations.StridedData

TensorOperations.StridedSubArray

TensorOperations.TensorOperations

TensorOperations.Zero

TensorOperations._dividebody

TensorOperations._filterdims

TensorOperations._findfirst

TensorOperations._findlast

TensorOperations._findnext

TensorOperations._indmax

TensorOperations._intersect

TensorOperations._isemptyset

TensorOperations._memjumps

TensorOperations._ncontree!

TensorOperations._one

TensorOperations._optimaltree

TensorOperations._permute

TensorOperations._scale!

TensorOperations._setdiff

TensorOperations._sreplace

TensorOperations._stridedloops

TensorOperations._union

TensorOperations._zero

TensorOperations.add! — Function.

add!(α, A, conjA, β, C, indCinA)

Implements C = β*C+α*permute(op(A)) where A is permuted according to indCinA and op is conj if conjA=Val{:C} or the identity map if conjA=Val{:N}. The indexable collection indCinA contains as nth entry the dimension of A associated with the nth dimension of C.

TensorOperations.add_indices

TensorOperations.add_micro!

TensorOperations.add_rec!

TensorOperations.add_strides

TensorOperations.axpby

TensorOperations.checkindices — Function.

checkindices(A, IA)

Checks whether the indices in IA are compatible with the object A. The fallback method checks length(IA) == numind(A).

TensorOperations.computecost

TensorOperations.connectedcomponents

TensorOperations.contract! — Function.

contract!(α, A, conjA, B, conjB, β, C, oindA, cindA, oindB, cindB, indCinoAB, [method])

Implements C = β*C+α*contract(op(A),op(B)) where A and B are contracted according to oindA, cindA, oindB, cindB and indCinoAB. The operation op acts as conj if conjA or conjB equal Val{:C} or as the identity map if conjA (conjB) equal Val{:N}. The dimension cindA[i] of A is contracted with dimension cindB[i] of B. The nth dimension of C is associated with an uncontracted (open) dimension of A or B according to indCinoAB[n] < NoA ? oindA[indCinoAB[n]] : oindB[indCinoAB[n]-NoA] with NoA=length(oindA) the number of open dimensions of A.

The optional argument method specifies whether the contraction is performed using BLAS matrix multiplication by specifying Val{:BLAS} (default), or using a native algorithm by specifying Val{:native}. The native algorithm does not copy the data but is typically slower.

TensorOperations.contract_indices

TensorOperations.contract_micro!

TensorOperations.contract_rec!

TensorOperations.contract_strides

TensorOperations.degree

TensorOperations.deindexify

TensorOperations.deindexify!

TensorOperations.eval

TensorOperations.getallindices

TensorOperations.getindices

TensorOperations.getlhsrhs

TensorOperations.hastraceindices

TensorOperations.include

TensorOperations.isassignment

TensorOperations.isdefinition

TensorOperations.isgeneraltensor

TensorOperations.isindex

TensorOperations.isnconstyle

TensorOperations.isscalarexpr

TensorOperations.istensor

TensorOperations.istensorexpr

TensorOperations.makegeneraltensor

TensorOperations.makeindex

TensorOperations.makescalar

TensorOperations.maketensor

TensorOperations.memjumps

TensorOperations.ncontree

TensorOperations.numind — Function.

numind(A)

Returns the number of indices of a tensor-like object A, i.e. for a multidimensional array (<:AbstractArray) we have numind(A) = ndims(A). Also works in type domain.

TensorOperations.optdata

TensorOperations.optimaltree

TensorOperations.parsecost

TensorOperations.prime

TensorOperations.processcontractorder

TensorOperations.rangeexpr

TensorOperations.scalar — Function.

scalar(C)

Returns the single element of a tensor-like object with zero dimensions, i.e. if numind(C)==0.

TensorOperations.similar_from_indices — Function.

similar_from_indices(::Type{T}, indices::NTuple{N,Int}, A, conjA=Val{:N}) where {T,N}

Returns an object similar to A which has an eltype given by T and dimensions/sizes corresponding to a selection of those of op(A), where the selection is specified by indices (which contains integer between 1 and numind(A)) and op is conj if conjA=Val{:C} or does nothing if conjA=Val{:N} (default).

similar_from_indices(::Type{T}, indices::NTuple{N,Int}, A, B, conjA=Val{:N}, conjB={:N}) where {T,N}

Returns an object similar to A which has an eltype given by T and dimensions/sizes corresponding to a selection of those of op(A) and op(B) concatenated, where the selection is specified by indices (which contains integers between 1 and numind(A)+numind(B) and op is conj if conjA or conjB equal Val{:C} or does nothing if conjA or conjB equal Val{:N} (default).

TensorOperations.storeset

TensorOperations.tensoradd — Function.

tensoradd(A, IA::Tuple, B, IB::Tuple, IC::Tuple=IA)

Returns the result of adding arrays A and B where the iterabels IA and IB denote how the array data should be permuted in order to be added. More specifically, the result of this method is equivalent to

tensorcopy(A, IA, IC) + tensorcopy(B, IB, IC)

but without creating the temporary permuted arrays.

TensorOperations.tensoradd!

TensorOperations.tensorcontract — Function.

tensorcontract(A, IA::Tuple, B, IB::Tuple, IC::Tuple; method=:BLAS)

Contract indices of array A with corresponding indices in array B by assigning them identical labels in the iterables IA and IB. The indices of the resulting array correspond to the indices that only appear in either IA or IB and can be ordered by specifying the optional argument IC. The default is to have all open indices of array A followed by all open indices of array B. Note that inner contractions of an array should be handled first with tensortrace, so that every label can appear only once in labelsA or labelsB seperately, and once (for open index) or twice (for contracted index) in the union of labelsA and labelsB.

The keyword argument method can take the values :BLAS or :native. The first option creates temporary copies of A, B and the resulting tensor where the indices are permuted such that the contractions become equivalent to a single matrix multiplication, which is then handled by BLAS. This is often the fastest approach and therefore the default value, but it does require sufficient memory and there is some overhead in allocating new memory (e.g. when doing this many times in a loop). In case method is set to :native, a native Julia function is called that performs the contraction without creating tempories, with special attention to cache-friendliness for maximal efficiency. This can be faster for small tensors or when the contraction pattern is non-generic, e.g. when it amounts to a scalar (no open indices) or a tensor product (no contraction indices).

TensorOperations.tensorcontract!

TensorOperations.tensorcopy — Function.

tensorcopy(A, IA, IC=IA)

Creates a copy of A, where the dimensions of A are assigned indices from the iterable IA and the indices of the copy are contained in IC. Both iterables should contain the same elements in a different order.

The result of this method is equivalent to permutedims(A, p) where p is the permutation such that IC=IA[p]. The implementation of tensorcopy is however more efficient on average.

TensorOperations.tensorcopy!

TensorOperations.tensorify

TensorOperations.tensorproduct — Function.

tensorproduct(A, IA::Tuple, B, IB::Tuple, IC::Tuple = (IA..., IB...))

Computes the tensor product of two arrays A and B, i.e. returns a new array C with ndims(C)=ndims(A)+ndims(B). The indices of the output tensor are related to those of the input tensors by the pattern specified by the indices. Essentially, this is a special case of tensorcontract with no indices being contracted over. On top of just calling tensorcontract with method=:native, this method checks whether the indices indeed specify a tensor product instead of a genuine contraction.

TensorOperations.tensorproduct!

TensorOperations.tensortrace — Function.

tensortrace(A, IA::Tuple, IC::Tuple)

Trace or contract pairs of indices of array A, by assigning them an identical indices in the iterable IA. The untraced indices, which are assigned a unique index, can be reordered according to the optional argument IC. The default value corresponds to the order in which they appear. Note that only pairs of indices can be contracted, so that every index in IA can appear only once (for an untraced index) or twice (for an index in a contracted pair).

TensorOperations.tensortrace!

TensorOperations.tinvperm

TensorOperations.trace! — Function.

trace!(α, A, conjA, β, C, indCinA, cindA1, cindA2)

Implements C = β*C+α*partialtrace(op(A)) where A is permuted and partially traced, according to indCinA, cindA1 and cindA2, and op is conj if conjA=Val{:C} or the identity map if conjA=Val{:N}. The indexable collection indCinA contains as nth entry the dimension of A associated with the nth dimension of C. The partial trace is performed by contracting dimension cindA1[i] of A with dimension cindA2[i] of A for all i in 1:length(cindA1).

TensorOperations.trace_indices

TensorOperations.trace_micro!

TensorOperations.trace_rec!

TensorOperations.trace_strides

TensorOperations.tree2expr

TensorOperations.tsetdiff

TensorOperations.tunique

TensorOperations.unique2 — Function.

unique2(itr)

Returns an array containing only those elements that appear exactly once in itr, and without any elements that appear more than once.