Docstrings

TimeseriesPrediction.@SMatrix

TimeseriesPrediction.@SVector

TimeseriesPrediction.@pack_PredictionParameters

TimeseriesPrediction.@pack_PredictionParameters!

TimeseriesPrediction.@unpack_PredictionParameters

TimeseriesPrediction.ASE

TimeseriesPrediction.AbstractBoundaryCondition — Type.

AbstractBoundaryCondition

Super-type of boundary conditions for SpatioTemporalEmbedding. Use subtypes(AbstractBoundaryCondition) for available methods.

TimeseriesPrediction.AbstractDataset

TimeseriesPrediction.AbstractEmbedding

TimeseriesPrediction.AbstractLocalModel — Type.

AbstractLocalModel

Supertype of methods for making a prediction of a query point q using local models, following the methods of [1]. Concrete subtypes are AverageLocalModel and LinearLocalModel.

All models weight neighbors with a chosen function, so that distant neighbors have smaller impact on the prediction and so that the interpolation is smooth. The default weighting function we use is

\[\begin{aligned} ω_i(d_i,d_{max}) = \left[ 1- \left(\frac{d_i}{d_{max}}\right)^2\right]^4 \end{aligned}\]

with $d_i = ||x_{nn,i} -q||_2$ being the distance of each neighbor from the query point.

You can also provide your own function or give ω_safe(d, dmax) = dmax > 0 ? (1.1 - (d/dmax)^2)^4 : 1.0 for a safe version of $ω$ that takes into acount edge cases. Finally you can also give nothing in place of ω. In that case no weighting is done and direct average of neighbors is returned.

Average Local Model

AverageLocalModel(ω)

The prediction is simply the weighted average of the images $y_{nn, i}$ of the neighbors $x_{nn, i}$ of the query point q, weighting using given function ω

\[\begin{aligned} y_{pred} = \frac{\sum{\omega_i y_{nn,i}}}{\sum{\omega_i}} \end{aligned}\]

Linear Local Model

LinearLocalModel([ω ], μ::Real=2.])
LinearLocalModel([ω ], s_min::Real, s_max::Real)

The prediction is a weighted linear regression over the neighbors $x_{nn, i}$ of the query and their images $y_{nn,i}$ as shown in [1].

Giving either μ or s_min and s_max determines which type of regularization is applied.

μ : Ridge Regression
\[\begin{aligned} f(\sigma) = \frac{\sigma^2}{\mu^2 + \sigma^2} \end{aligned}\]

s_min, s_max : Soft Threshold
\[\begin{aligned} f(\sigma) = \begin{cases} 0, & \sigma < s_{min}\\ \left(1 - \left( \frac{s_{max}-\sigma}{s_{max}-s_{min}}\right)^2 \right)^2, &s_{min} \leq \sigma \leq s_{max} \\ 1, & \sigma > s_{max}\end{cases} \end{aligned}\]

References

[1] : D. Engster & U. Parlitz, Handbook of Time Series Analysis Ch. 1, VCH-Wiley (2006)

TimeseriesPrediction.AbstractNeighborhood

TimeseriesPrediction.AbstractSpatialEmbedding — Type.

AbstractSpatialEmbedding <: AbstractEmbedding

Super-type of spatiotemporal embedding methods. Valid subtypes:

SpatioTemporalEmbedding

PCAEmbedding

TimeseriesPrediction.AverageLocalModel — Type.

AverageLocalModel(ω::Function = ω_unsafe)

See AbstractLocalModel.

TimeseriesPrediction.CDS_KWARGS

TimeseriesPrediction.Chebyshev

TimeseriesPrediction.Cityblock

TimeseriesPrediction.ConstantBoundary — Type.

ConstantBoundary(c) <: AbstractBoundaryCondition

Constant boundary condition type. Enforces constant boundary conditions when passed to SpatioTemporalEmbedding by filling missing out-of-bounds values in the reconstruction with parameter c.

TimeseriesPrediction.ContinuousDynamicalSystem

TimeseriesPrediction.Dataset

TimeseriesPrediction.DelayEmbedding

TimeseriesPrediction.DelayEmbeddings

TimeseriesPrediction.DiscreteDynamicalSystem

TimeseriesPrediction.DynamicalSystem

TimeseriesPrediction.DynamicalSystemsBase

TimeseriesPrediction.Euclidean

TimeseriesPrediction.FixedMassNeighborhood

TimeseriesPrediction.FixedSizeNeighborhood

TimeseriesPrediction.KDTree

TimeseriesPrediction.LinearLocalModel — Type.

LinearLocalModel([ω ], μ::Real=2.])
LinearLocalModel([ω ], s_min::Real, s_max::Real)

See AbstractLocalModel.

TimeseriesPrediction.MSE1

TimeseriesPrediction.MSEp — Function.

MSEp(R::AbstractDataset{D,T}, R_test, p; method, ntype, stepsize) -> error

Compute mean squared error of iterated predictions of length p using test set R_test.

Description

This error measure takes in a prediction model consisting of R, method, ntype and stepsize and evaluates its performance. The test set R_test is a delay reconstruction with the same delay τ and dimension D as R. For each subset of R_test with length p it calls localmodel_tsp. The model error is then defined as

\[\begin{aligned} MSE_p = \frac{1}{p|T_{ref}|}\sum_{t\in T_{ref}}\sum_{i=1}^{p} \left(y_{t+i} - y_{pred,t+i} \right)^2 \end{aligned}\]

where $|T_{ref}|$ is the number of subsets of R_test used.

References

See localmodel_tsp.

TimeseriesPrediction.MTDelayEmbedding

TimeseriesPrediction.PCAEmbedding — Type.

PCAEmbedding(s, em::SpatioTemporalEmbedding; kwargs...) → embedding

A spatio temporal delay coordinates structure with Principal Component Analysis as a means of dimension reduction, embedding can be used as a functor:

embedding(rvec, s, t, α)

which operates inplace on rvec and reconstructs vector from spatial time series s at timestep t and cartesian index α.

To instantiate this embedding, give the data to be reconstructed s as well as an instance of SpatioTemporalEmbedding to PCAEmbedding.

Keyword Arguments

pratio = 0.99 : Ratio of variances that needs to be preserved in low-dimensional PCA-reconstruction.

maxoutdim = 25: Upper limit for output dimension. May break pratio criterion.

every_t = 1 : Speed up computation by only using every n-th point in time.

every_α = 1 : Speed up computation further by only using every n-th point in space (linear indexing).

To set the output dimension to a certain value X, pass pratio=1, maxoutdim=X.

TimeseriesPrediction.PeriodicBoundary — Type.

PeriodicBoundary <: AbstractBoundaryCondition

Periodic boundary condition struct. Enforces periodic boundary conditions when passed to SpatioTemporalEmbedding in the reconstruction.

TimeseriesPrediction.PredictionParameters

TimeseriesPrediction.Region — Type.

Region{Φ}

Internal struct for efficiently keeping track of region far from boundaries of field. Used to speed up reconstruction process.

TimeseriesPrediction.SMatrix

TimeseriesPrediction.STE

TimeseriesPrediction.SVector

TimeseriesPrediction.SpatioTemporalEmbedding — Type.

SpatioTemporalEmbedding{T,Φ,BC,X} → embedding

A spatio temporal delay coordinates structure to be used as a functor. Applies to data of Φ spatial dimensions and gives an embedding of dimensionality X.

embedding(rvec, s, t, α)

Operates inplace on rvec (of length X) and reconstructs vector from spatial timeseries s at timestep t and cartesian index α. Note that there are no bounds checks for t.

It is assumed that s is a Vector{<:AbstractArray{T,Φ}}.

Constructors

There are some convenience constructors that return intuitive embeddings here:

cubic_shell_embedding

light_cone_embedding

The "main" constructor is

SpatioTemporalEmbedding{X}(τ, β, bc, fsize)

which allows full control over the spatio-temporal embedding.

Χ == length(τ) == length(β) : dimensionality of resulting reconstructed space.

τ::Vector{Int} = Vector of temporal delays for each entry of the reconstructed space (sorted in ascending order).

β::Vector{CartesianIndex{Φ}} = vector of relative indices of spatial delays for each entry of the reconstructed space.

bc::BC : boundary condition.

fsize::NTuple{Φ, Int} : Size of each state in the timeseries.

TimeseriesPrediction.Systems

TimeseriesPrediction.TimeseriesPrediction — Module.

Prediction of timeseries using methods of nonlinear dynamics and timeseries analysis

TimeseriesPrediction._localmodel_tsp

TimeseriesPrediction.autocor

TimeseriesPrediction.columns

TimeseriesPrediction.compute_pca

TimeseriesPrediction.convert_idx

TimeseriesPrediction.crossprediction — Function.

crossprediction(source_train, target_train, source_pred,
                em::AbstractSpatialEmbedding; kwargs...)

Perform a spatio-temporal timeseries cross-prediction for target from source, using local weighted modeling [1]. This can be used for example when there are coupled spatial fields and one is used to predict the other. It is assumed that source_train, target_train, source_pred are all of the same type, AbstractVector{<:AbstractArray{T, Φ}}.

The spatio temporal delay embedding process is defined by em. See AbstractSpatialEmbedding for available methods and interfaces.

Keyword Arguments

ttype = KDTree : Type/Constructor of tree structure. So far only tested with KDTree.

method = AverageLocalModel(ω_safe) : Subtype of AbstractLocalModel.

ntype = FixedMassNeighborhood(3) : Subtype of AbstractNeighborhood.

progress = true : To print progress done.

Description

The reconstructed state of source_train[t][i,j,...] is associated with the output value target_train[t][i,j,...]. This establishes a "connection" between target and source. Taking a reconstructed state of source_pred as query point, the function finds its neighbors in the reconstructed space of source_train using neighborhood ntype. Then, the neighbor indices are used to make a prediction for the corresponding value of the target, using the established "connection" between fields.

Additional Interfaces

To save computation time in the case of repeated predictions with the same training set and embedding parameters we provide an additional interface that allows the user to provide an existing reconstruction and tree structure.

R = reconstruct(train_in, em)
tree = ttype(R)
params = PredictionParams(em, method, ntype, ttype)
sol = crossprediction(params, train_out, pred_in, R, tree; progress=true)

where params is an internal container with all relevant parameters.

Performance Notes

Be careful when choosing embedding parameters as memory usage and computation time depend strongly on the resulting embedding dimension.

References

[1] : U. Parlitz & C. Merkwirth, Phys. Rev. Lett. 84, pp 1890 (2000)

TimeseriesPrediction.cubic_shell_embedding — Function.

cubic_shell_embedding(s, D, τ, B, k, bc) → embedding

Create a SpatioTemporalEmbedding instance that includes spatial neighbors in hypercubic shells. The embedding is to be used with data from s.

Description

Points are participating in the embedding by forming hypercubic shells around the current point. The total shells formed are B. The points on the shells have spatial distance k ≥ 1 (distance in indices, like a cityblock metric). k = 1 means that all points of the shell participate. The points of the hypercubic grid can be separated by k ≥ 1 points apart (i.e. dropping k-1 in-between points). In short, in each spatial dimension of the system the cartesian offset indices are -B*k : k : k*B.

D is the number of temporal steps in the past to be included in the embedding, where each step in the past has additional delay time τ::Int. D=0 corresponds to using only the present. Notice that all embedded time frames have the same spatial structure, in contrast to light_cone_embedding.

As an example, consider one of the D embedded frames (all are the same) of a system with 2 spatial dimensions (□ = current point, (included by definition in the embedding), n = included points in the embedding coming from n-th shell, . = points not included in the embedding)

      B = 2,  k = 1        |        B = 1,  k = 2        |        B = 2,  k = 2
                           |                             |
.  .  .  .  .  .  .  .  .  |  .  .  .  .  .  .  .  .  .  |  2  .  2  .  2  .  2  .  2
.  .  .  .  .  .  .  .  .  |  .  .  .  .  .  .  .  .  .  |  .  .  .  .  .  .  .  .  .
.  .  2  2  2  2  2  .  .  |  .  .  1  .  1  .  1  .  .  |  2  .  1  .  1  .  1  .  2
.  .  2  1  1  1  2  .  .  |  .  .  .  .  .  .  .  .  .  |  .  .  .  .  .  .  .  .  .
.  .  2  1  □  1  2  .  .  |  .  .  1  .  □  .  1  .  .  |  2  .  1  .  □  .  1  .  2
.  .  2  1  1  1  2  .  .  |  .  .  .  .  .  .  .  .  .  |  .  .  .  .  .  .  .  .  .
.  .  2  2  2  2  2  .  .  |  .  .  1  .  1  .  1  .  .  |  2  .  1  .  1  .  1  .  2
.  .  .  .  .  .  .  .  .  |  .  .  .  .  .  .  .  .  .  |  .  .  .  .  .  .  .  .  .
.  .  .  .  .  .  .  .  .  |  .  .  .  .  .  .  .  .  .  |  2  .  2  .  2  .  2  .  2

TimeseriesPrediction.cut_off_beginning!

TimeseriesPrediction.dimension

TimeseriesPrediction.embed

TimeseriesPrediction.estimate_delay

TimeseriesPrediction.estimate_dimension

TimeseriesPrediction.eval

TimeseriesPrediction.exponential_decay_fit

TimeseriesPrediction.gen_queries

TimeseriesPrediction.get_deviations

TimeseriesPrediction.get_num_pt

TimeseriesPrediction.get_state

TimeseriesPrediction.get_usable_idxs

TimeseriesPrediction.get_τmax

TimeseriesPrediction.include

TimeseriesPrediction.indices_within_sphere — Function.

indices_within_sphere(r, Φ) → β

Return all cartesian indices within a hypersphere or radius r and dimension Φ.

TimeseriesPrediction.inner_region

TimeseriesPrediction.integrator

TimeseriesPrediction.jacobian

TimeseriesPrediction.light_cone_embedding — Function.

light_cone_embedding(s, D, τ, r₀, c, bc) → embedding

Create a SpatioTemporalEmbedding instance that includes spatial and temporal neighbors of a point based on the notion of a light cone.

The embedding is to be used with data from s.

Description

Information does not travel instantly but with some finite speed c ≥ 0.0. This constructor creates a cone-like embedding including all points in space and time, whose value can influence a prediction based on the information speed c. D is the number of temporal steps in the past to be included in the embedding, where each step in the past has additional delay time τ::Int. D=0 corresponds to using only the present. r₀ is the initial radius at the top of the cone, i.e. the radius of influence at the present. bc is the boundary condition.

The radius of the light cone evolves as: r = i*τ*c + r₀ for each step i ∈ 0:D.

As an example, in a one-dimensional system with D = 1, τ = 2, r₀ = 1, the embedding looks like (□ = current point (included by definition in the embedding), o point to be predicted using temporalprediction, x = points included in the embedding, . = points not included in the embedding)

time  | c = 1.0               | c = 2.0               | c = 0.0

n + 1 | ..........o.......... | ..........o.......... | ..........o..........
n     | .........x□x......... | .........x□x......... | .........x□x.........
n - 1 | ..................... | ..................... | .....................
n - τ | .......xxxxxxx....... | .....xxxxxxxxxx...... | .........xxx.........

Besides this example, in the official documentation we show a function explain_light_cone which produces a plot of the light cone for 2 spatial dimensions (great for understanding!).

TimeseriesPrediction.localmodel_cp — Function.

localmodel_cp(source_pool, target_pool, source_pred,  D, τ; kwargs...)

Perform a cross prediction from source to target, using local weighted modeling [1]. source_pred is the input for the prediction and source_pool and target_pool are used as pooling/training data for the predictions. The function always returns an object of the same type as target_pool, which can be either a timeseries (vector) or an AbstractDataset (trajectory).

Keyword Arguments

method = AverageLocalModel(ω_unsafe) : Subtype of AbstractLocalModel.

ntype = FixedMassNeighborhood(2) : Subtype of AbstractNeighborhood.

stepsize = 1 : Prediction step size.

Instead of passing D & τ for reconstruction one may also give existing Datasets as source_pool and source_pred. In this case an additional keyword argument y_idx_shift::Int=0 may be necessary to account for the index shift introduced in the reconstruction process.

Description

Given a query point, the function finds its neighbors using neighborhood ntype. Then, the neighbors xnn and their images ynn are used to make a prediction for the image of the query point, using the provided method.

References

[1] : D. Engster & U. Parlitz, Handbook of Time Series Analysis Ch. 1, VCH-Wiley (2006)

TimeseriesPrediction.localmodel_tsp — Function.

localmodel_tsp(s, D::Int, τ, p::Int; method, ntype, stepsize)
localmodel_tsp(s, p::Int; method, ntype, stepsize)

Perform a timeseries prediction for p points, using local weighted modeling [1]. The function always returns an object of the same type as s, which can be either a timeseries (vector) or an AbstractDataset (trajectory), and the returned data always contains the final point of s as starting point. This means that the returned data has length of p + 1.

If given (s, D, τ), it first calls reconstruct on s with dimension D and delay τ. If given only s then no reconstruction is done.

Keyword Arguments

method = AverageLocalModel(ω_unsafe) : Subtype of AbstractLocalModel.

ntype = FixedMassNeighborhood(2) : Subtype of AbstractNeighborhood.

stepsize = 1 : Prediction step size.

Description

Given a query point, the function finds its neighbors using neighborhood ntype. Then, the neighbors xnn and their images ynn are used to make a prediction for the future of the query point, using the provided method. The images ynn are the points xnn shifted by stepsize into the future.

The algorithm is applied iteratively until a prediction of length p has been created, starting with the query point to be the last point of the timeseries.

References

[1] : D. Engster & U. Parlitz, Handbook of Time Series Analysis Ch. 1, VCH-Wiley (2006)

TimeseriesPrediction.maxima

TimeseriesPrediction.mcnames_reg

TimeseriesPrediction.min_pairwise_distance

TimeseriesPrediction.minima

TimeseriesPrediction.minmaxima

TimeseriesPrediction.mutualinformation

TimeseriesPrediction.neighborhood

TimeseriesPrediction.neighborhood_and_distances

TimeseriesPrediction.orthonormal

TimeseriesPrediction.outdim

TimeseriesPrediction.parallel_integrator

TimeseriesPrediction.project_inside

DelayEmbeddings.reconstruct — Function.

reconstruct(s::AbstractArray{<:AbstractArray{T,Φ}}, em)

Reconstruct the spatial timeseries s represented by a Vector of AbstractArray states using the embedding struct em of type AbstractSpatialEmbedding.

Returns the reconstruction in the form of a Dataset where each row is a reconstructed state and they are ordered first through linear indexing into each state and then incrementing in time.

TimeseriesPrediction.reinit!

TimeseriesPrediction.ridge_reg

TimeseriesPrediction.set_deviations!

TimeseriesPrediction.set_parameter!

TimeseriesPrediction.set_state!

TimeseriesPrediction.step!

TimeseriesPrediction.stochastic_indicator

TimeseriesPrediction.tangent_integrator

TimeseriesPrediction.temporalprediction — Function.

temporalprediction(U, em::AbstractSpatialEmbedding, tsteps; kwargs...)

Perform a spatio-temporal time series prediction for tsteps iterations, using local weighted modeling [1] give a time series of the form U::AbstractVector{<:AbstractArray{T, Φ}}.

The returned data always contains the final state of U as starting point (total returned length is tsteps+1). The reconstruction process is defined by em. For available methods and interfaces see AbstractSpatialEmbedding.

Keyword Arguments

ttype = KDTree : Type/Constructor of tree structure. So far only tested with KDTree.

method = AverageLocalModel(ω_safe) : Subtype of AbstractLocalModel.

ntype = FixedMassNeighborhood(3) : Subtype of AbstractNeighborhood.

initial_ts = U : Initial states for prediction (same type as U). Must have at least as many states as the maximum delay time used. Defaults to the training set U.

progress = true : To print progress done.

Description

This method works similarly to localmodel_tsp, by expanding the concept of delay embedding to spatially extended systems. Instead of reconstructing complete states of the system, local states are used. See AbstractSpatialEmbedding for details on the embedding. Predictions are then performed frame by frame and point py point. Once all values for a new frame are found, the frame is added to the end of the timeseries and used to generate new prediction queries for the next time step.

Performance Notes

Be careful when choosing embedding parameters as memory usage and computation time depend strongly on the resulting embedding dimension.

References

[1] : U. Parlitz & C. Merkwirth, Phys. Rev. Lett. 84, pp 1890 (2000)

TimeseriesPrediction.trajectory

TimeseriesPrediction.working_ts

TimeseriesPrediction.ω_safe

TimeseriesPrediction.ω_unsafe