Docstrings

DeferredFutures.@defer — Macro.

@defer Future(...)
@defer RemoteChannel(...)

@defer transforms a Future or RemoteChannel construction into a 'DeferredFuture' or 'DeferredChannel' construction.

DeferredFutures.DeferredChannel — Type.

DeferredChannel(pid::Integer=myid(), num::Integer=1; content::DataType=Any) -> DeferredChannel

Create a DeferredChannel with a reference to a remote channel of a specific size and type. f() is a function that when executed on pid must return an implementation of an AbstractChannel.

The default pid is the current process.

DeferredChannel(pid::Integer=myid(), num::Integer=1; content::DataType=Any) -> DeferredChannel

Create a DeferredChannel. The default pid is the current process. When initialized, the DeferredChannel will reference a Channel{content}(num) on process pid.

Note that the data in the DeferredChannel will still be located wherever the first piece of data was put! from. The pid argument controls where the outermost reference to that data is located.

DeferredFutures.DeferredFuture — Type.

DeferredFuture(pid::Integer=myid()) -> DeferredFuture

Create a DeferredFuture on process pid. The default pid is the current process.

Note that the data in the DeferredFuture will still be located wherever it was put! from. The pid argument controlls where the outermost reference to that data is located.

DeferredFutures.DeferredFutures

DeferredFutures.DeferredRemoteRef — Type.

DeferredRemoteRef is the common supertype of DeferredFuture and DeferredChannel and is the counterpart of Distributed.AbstractRemoteRef.

DeferredFutures.eval

DeferredFutures.finalize_ref — Function.

finalize_ref(ref::DeferredRemoteRef)

This finalizer is attached to both DeferredFuture and DeferredChannel on construction and finalizes the inner and outer RemoteChannels.

For more information on finalizing remote references, see the Julia manual[1].

[1]
Remote References and Distributed Garbage Collection

DeferredFutures.include

DeferredFutures.reset! — Function.

reset!{T<:DeferredRemoteRef}(ref::T) -> T

Removes any data from the DeferredRemoteRef and allows it to be reinitialized with data.

Returns the input DeferredRemoteRef.