Docstrings

RData.ANYSXP

RData.ASCIIIO — Type.

ASCII RData format IO stream wrapper.

RData.ATTRLANGSXP

RData.ATTRLISTSXP

RData.BASEENV_SXP

RData.BASENAMESPACE_SXP

RData.BCODESXP

RData.BCREPDEF

RData.BCREPREF

RData.BUILTINSXP

RData.BYTECODELANG_Types

RData.BytecodeContext — Type.

Context for reading R bytecode.

RData.CHARSXP

RData.CHARSXProps

RData.CLASSREFSXP

RData.CLOSXP

RData.CPLXSXP

RData.DOTSXP

RData.DictoVec — Type.

Container that mimics R vector behaviour. Elements could be accessed either by indices as a normal vector, or (optionally) by string keys as a dictionary.

RData.EMPTYENV_SXP

RData.ENVSXP

RData.EXPRSXP

RData.EXTPTRSXP

RData.FREESXP

RData.FUNSXP

RData.GENERICREFSXP

RData.GLOBALENV_SXP

RData.Hash

RData.INTSXP

RData.LANGSXP

RData.LGLSXP

RData.LISTSXP

RData.LONG_VECTOR_SUPPORT

RData.MISSINGARG_SXP

RData.NAMESPACESXP

RData.NEWSXP

RData.NILSXP

RData.NILVALUE_SXP

RData.NativeIO — Type.

Native binary RData format IO stream wrapper.

TODO write readers

RData.PACKAGESXP

RData.PERSISTSXP

RData.PROMSXP

RData.RAWSXP

RData.RBuiltin

RData.RBytecode

RData.RClosure

RData.RComplexVector — Type.

R vector object.

RData.RDAContext — Type.

RDA (R data archive) reading context.

Stores flags that define how R objects are read and converted into Julia objects.

Maintains the list of R objects that could be referenced later in the RDA stream.

RData.RDAIO — Type.

Abstract RDA format IO stream wrapper.

RData.RDATag

RData.RData

RData.RDummy

RData.REALSXP

RData.REFSXP

RData.REnvironment

RData.RExtPtr

RData.RIntegerVector — Type.

R vector object.

RData.RList — Type.

R vector object.

RData.RLogicalVector — Type.

R vector object.

RData.RNamespace

RData.RNullableVector — Type.

R vector object with explicit NA values.

RData.RNumericVector — Type.

R vector object.

RData.ROBJ — Type.

Base class for all R types (objects) that can have attributes.

RData.RPackage

RData.RPairList — Type.

Representation of R's paired list-like structures (LISTSXP, LANGSXP). Unlike R which represents these as singly-linked list, RPairList uses vector representation.

RData.RPromise

RData.RRaw

RData.RS4Obj

RData.RSEXPREC — Type.

Base class for RData internal representation of all R types. SEXPREC stands for S (R predecessor) expression record.

RData.RString

RData.RStringVector — Type.

R vector object with explicit NA values.

RData.RSymbol — Type.

R symbol. Not quite the same as a Julia symbol.

RData.RVEC — Type.

Base class for all R vector-like objects.

RData.RVecLength

RData.RVector — Type.

R vector object.

RData.R_Date_Class

RData.R_NA_FLOAT64

RData.R_NA_FLOAT64_LOW — Constant.

Non-standard addition to NaN bit pattern to discriminate NA from NaN. 0x000007a2 == UInt32(1954) (I assume 1954 is the year of Ross's birth or something like that.)

RData.R_NA_INT32

RData.R_NA_STRING

RData.R_POSIXct_Class

RData.S4SXP

RData.SPECIALSXP

RData.STRSXP

RData.SXType

RData.SXTypeInfo — Type.

Definition of R type.

RData.SXTypes — Constant.

Maps R type id (SXType) to its SXTypeInfo.

RData.SYMSXP

RData.UNBOUNDVALUE_SXP

RData.VECSXP

RData.WEAKREFSXP

RData.XDRIO — Type.

XDR (machine-independent binary) RData format IO stream wrapper.

RData._unix2zdt

RData.class

RData.decompress

RData.emptyhash

RData.emptyhashkey

RData.emptystrvec

RData.eval

RData.fileio_load

RData.getattr

RData.getjuliatz

RData.hasattr

RData.hasdim

RData.hasdimnames

RData.hasnames

RData.hastag

RData.include

RData.inherits

RData.int2ver

RData.isdataframe

RData.isfactor

RData.isna

RData.isna_float64

RData.isobj

RData.jlvec

RData.load

RData.na2zero

RData.name2index

RData.r2juliatz

RData.rdaio — Function.

Creates RDAIO wrapper for io stream depending on its format specified by formatcode.

RData.rdays2date

RData.readS4

RData.readattrs

RData.readbuiltin

RData.readbytecode

RData.readbytecodeconsts

RData.readbytecodecontents

RData.readbytecodelang

RData.readcharacter

RData.readcharsxprops

RData.readclosure

RData.readcomplex

RData.readdummy

RData.readenv

RData.readextptr

RData.readfloat64

RData.readfloatorNA

RData.readfloatorNA!

RData.readint32

RData.readinteger

RData.readintorNA

RData.readitem

RData.readlang

RData.readlength

RData.readlist

RData.readlogical

RData.readname

RData.readnamespace

RData.readnchars

RData.readnil

RData.readnumeric

RData.readpackage

RData.readpairedobjects

RData.readpairlist

RData.readpromise

RData.readraw

RData.readref

RData.readstring

RData.readsymbol

RData.readuint32

RData.readuint8

RData.readunsupported

RData.registerref! — Function.

Register R object, so that it could be referenced later (by its index in the reference table).

RData.row_names

RData.sexp2julia

RData.sxtype