Docstrings

Mustache.@mt_mstr

Mustache.@mt_str — Macro.

mt"string"

Macro to parse tokens from a string. Useful when template is to be reused.

Mustache.AnIndex

Mustache.Context

Mustache.Mustache

Mustache.MustacheTokens

Mustache.SectionToken

Mustache.TEMPLATES

Mustache.TagToken

Mustache.TextToken

Mustache.Token

Mustache.WHITESPACE

Mustache.Writer

Mustache._lookup_in_view

Mustache._renderTokensByValue

Mustache._toString

Mustache.asRegex

Mustache.clearCache

Mustache.compile

Mustache.compilePartial

Mustache.compileTokens

Mustache.ctx_push

Mustache.curlyRe

Mustache.end_of_road

Mustache.entityMap

Mustache.eqRe

Mustache.escapeRe

Mustache.escapeTags

Mustache.escape_html

Mustache.eval

Mustache.falsy

Mustache.getPartial

Mustache.include

Mustache.isWhitespace

Mustache.is_dataframe

Mustache.is_l_standalone

Mustache.is_r_standalone

Mustache.lookup

Mustache.lookup_dotted

Mustache.lookup_in_view

Mustache.make_tokens

Mustache.nestTokens

Mustache.nonSpaceRe

Mustache.parse

Mustache.peekaheadmatch

Mustache.peekchar

Mustache.popfirst!_whitespace

Mustache.readutf

Mustache.render — Function.

render([io], tokens, view)

Render a set of tokens with a view, using optional io object to print or store.

Arguments

io::IO: Optional IO object.

tokens: Either Mustache tokens, or a string to parse into tokens

view: A view provides a context to look up unresolved symbols demarcated by mustache braces. A view may be specified by a dictionary, a module, a composite type, a vector, a named tuple, a data frame, a Tables object, or keyword arguments.

Mustache.renderTokens

Mustache.renderTokensByValue

Mustache.render_from_file — Function.

Renders a template from filepath and view. If it has seen the file before then it finds the compiled MustacheTokens in TEMPLATES rather than calling parse a second time.

Mustache.scan_past!

Mustache.scan_until!

Mustache.spaceRe

Mustache.stripWhitespace

Mustache.tagRe

Mustache.template_from_file

Mustache.toString

Mustache.utf8_offset

Mustache.utf8_trailing

Mustache.whiteRe