Docstrings

WeakRefStrings.MISSING_OFFSET

WeakRefStrings.NULLSTRING

WeakRefStrings.NULLSTRING16

WeakRefStrings.NULLSTRING32

WeakRefStrings.STR

WeakRefStrings.StringArray — Type.

StringArray{T,N}

Efficient storage for N dimensional array of strings.

StringArray stores underlying string data for all elements of the array in a single contiguous buffer. It maintains offset and length for each element.

T can be String, WeakRefString, Union{Missing, String} or Union{Missing, WeakRefString}. getindex will return this type although all variants have the same storage format.

You can use convert(StringArray{U}, ::StringArray{T}) to change the element type (e.g. to WeakRefString for efficiency) without copying the data.

Example construction

julia> sa = StringArray(["x", "y"]) # from Array{String}
2-element WeakRefStrings.StringArray{String,1}:
 "x"
 "y"

julia> sa = StringArray{WeakRefString}(["x", "y"])
2-element WeakRefStrings.StringArray{WeakRefStrings.WeakRefString,1}:
 "x"
 "y"

julia> sa = StringArray{Union{Missing, String}}(["x", "y"]) # with Missing
2-element WeakRefStrings.StringArray{Union{Missings.Missing, String},1}:
 "x"
 "y"

julia> sa = StringArray{Union{Missing, String}}(2,2) # undef
2×2 WeakRefStrings.StringArray{Union{Missings.Missing, String},2}:
 #undef  #undef
 #undef  #undef

WeakRefStrings.StringVector

WeakRefStrings.UNDEF_OFFSET

WeakRefStrings.WeakRefString — Type.

A custom "weakref" string type that only points to external string data. Allows for the creation of a "string" instance without copying data, which allows for more efficient string parsing/movement in certain data processing tasks.

Please note that no original reference is kept to the parent string/memory, so WeakRefString becomes unsafe once the parent object goes out of scope (i.e. loses a reference to it)

Internally, a WeakRefString{T} holds:

ptr::Ptr{T}: a pointer to the string data (code unit size is parameterized on T)

len::Int: the number of code units in the string data

See also WeakRefStringArray

WeakRefStrings.WeakRefStringArray — Type.

A WeakRefString container. Holds the "strong" references to the data pointed by its strings, ensuring that the referenced memory blocks stay valid during WeakRefStringArray lifetime.

Upon indexing an elemnt in a WeakRefStringArray, the underlying WeakRefString is converted to a proper Julia String type by copying the memory; this ensures safe string processing in the general case. If additional optimizations are desired, the direct WeakRefString elements can be accessed by indexing A.elements, where A is a WeakRefStringArray.

WeakRefStrings.WeakRefStrings

WeakRefStrings._isassigned

WeakRefStrings._setindex!

WeakRefStrings.chompnull

WeakRefStrings.eval

WeakRefStrings.include

WeakRefStrings.init

WeakRefStrings.iterate_continued

WeakRefStrings.wk