Docstrings

PlotUtils.@colorant_str

PlotUtils.ABGR

PlotUtils.ADIN99

PlotUtils.ADIN99d

PlotUtils.ADIN99o

PlotUtils.AGray

PlotUtils.AGray32

PlotUtils.AHSI

PlotUtils.AHSL

PlotUtils.AHSV

PlotUtils.ALCHab

PlotUtils.ALCHuv

PlotUtils.ALMS

PlotUtils.ALab

PlotUtils.ALuv

PlotUtils.ARGB

PlotUtils.ARGB32

PlotUtils.AXYZ

PlotUtils.AYCbCr

PlotUtils.AYIQ

PlotUtils.AbstractGray

PlotUtils.AbstractRGB

PlotUtils.AlphaColor

PlotUtils.AxyY

PlotUtils.BGR

PlotUtils.BGRA

PlotUtils.CIE1931JV_CMF

PlotUtils.CIE1931J_CMF

PlotUtils.CIE1931_CMF

PlotUtils.CIE1964_CMF

PlotUtils.Color

PlotUtils.Color3

PlotUtils.ColorAlpha

PlotUtils.ColorGradient — Type.

Continuous gradient between values. Wraps a list of bounding colors and the values they represent.

PlotUtils.ColorLibrary

PlotUtils.ColorTypes

PlotUtils.Colorant

PlotUtils.ColorantNormed

PlotUtils.Colors

PlotUtils.DE_2000

PlotUtils.DE_94

PlotUtils.DE_AB

PlotUtils.DE_BFD

PlotUtils.DE_CMC

PlotUtils.DE_DIN99

PlotUtils.DE_DIN99d

PlotUtils.DE_DIN99o

PlotUtils.DE_JPC79

PlotUtils.DIN99

PlotUtils.DIN99A

PlotUtils.DIN99d

PlotUtils.DIN99dA

PlotUtils.DIN99o

PlotUtils.DIN99oA

PlotUtils.Fractional

PlotUtils.Gray

PlotUtils.Gray24

PlotUtils.GrayA

PlotUtils.HSB

PlotUtils.HSI

PlotUtils.HSIA

PlotUtils.HSL

PlotUtils.HSLA

PlotUtils.HSV

PlotUtils.HSVA

PlotUtils.LCHab

PlotUtils.LCHabA

PlotUtils.LCHuv

PlotUtils.LCHuvA

PlotUtils.LMS

PlotUtils.LMSA

PlotUtils.Lab

PlotUtils.LabA

PlotUtils.Luv

PlotUtils.LuvA

PlotUtils.MSC

PlotUtils.PlotUtils

PlotUtils.RGB

PlotUtils.RGB1

PlotUtils.RGB24

PlotUtils.RGB4

PlotUtils.RGBA

PlotUtils.Transparent3

PlotUtils.TransparentColor

PlotUtils.TransparentGray

PlotUtils.TransparentRGB

PlotUtils.XYZ

PlotUtils.XYZA

PlotUtils.YCbCr

PlotUtils.YCbCrA

PlotUtils.YIQ

PlotUtils.YIQA

PlotUtils.__init__

PlotUtils._color_list

PlotUtils._gradients

PlotUtils._lch_c_const

PlotUtils._lightness_darkbg

PlotUtils._lightness_lightbg

PlotUtils._rainbowColors

PlotUtils._testColors

PlotUtils.adapted_grid — Function.

adapted_grid(f, minmax::Tuple{Number, Number}; max_recursions = 7)

Computes a grid x on the interval [minmax[1], minmax[2]] so that plot(f, x) gives a smooth "nice" plot. The method used is to create an initial uniform grid (21 points) and refine intervals where the second derivative is approximated to be large. When an interval becomes "straight enough" it is no longer divided. Functions are never evaluated exactly at the end points of the intervals.

The parameter max_recusions computes how many times each interval is allowed to be refined.

PlotUtils.adjust_away

PlotUtils.adjust_lch

PlotUtils.alpha

PlotUtils.alphacolor

PlotUtils.base_color_type

PlotUtils.base_colorant_type

PlotUtils.blue

PlotUtils.bounding_order_of_magnitude

PlotUtils.ccolor

PlotUtils.cgrad

PlotUtils.cgrad_colors

PlotUtils.cgrad_reverse

PlotUtils.cgradients — Function.

cgradients([color_library::Symbol])

List available color gradients in color_library (defaults to the currently loaded library)

PlotUtils.cie_color_match

PlotUtils.clibraries — Function.

clibraries()

List the available color libraries on the system

PlotUtils.clibrary — Function.

clibrary(grad::Symbol)

Set the active color library. A list of possible libraries can be printed with clibraries()

PlotUtils.color

PlotUtils.color_libraries

PlotUtils.color_type

PlotUtils.coloralpha

PlotUtils.colordiff

PlotUtils.colormap

PlotUtils.colormatch

PlotUtils.comp1

PlotUtils.comp2

PlotUtils.comp3

PlotUtils.cvec

PlotUtils.darken

PlotUtils.default_cgrad

PlotUtils.deuteranopic

PlotUtils.distinguishable_colors

PlotUtils.diverging_palette

PlotUtils.eval

PlotUtils.gamutmax

PlotUtils.gamutmin

PlotUtils.generate_colorgradient

PlotUtils.get_color_palette

PlotUtils.get_zvalues

PlotUtils.getgradient

PlotUtils.getpctrange

PlotUtils.gradient_from_list

PlotUtils.gray

PlotUtils.green

PlotUtils.hex

PlotUtils.include

PlotUtils.interpolate

PlotUtils.interpolate_rgb

PlotUtils.invisible

PlotUtils.iscgrad_symbol

PlotUtils.isdark

PlotUtils.islight

PlotUtils.lighten

PlotUtils.lightness_from_background

PlotUtils.lightness_level

PlotUtils.make255

PlotUtils.mapc

PlotUtils.mapreducec

PlotUtils.multilevel_ticks

PlotUtils.optimize_datetime_ticks

PlotUtils.optimize_ticks

PlotUtils.optimize_ticks_typed

PlotUtils.plot_color

PlotUtils.protanopic

PlotUtils.red

PlotUtils.reducec

PlotUtils.register_color_library

PlotUtils.register_gradient_colors

PlotUtils.rgb_string

PlotUtils.rgba_string

PlotUtils.sample_evenly

PlotUtils.sequential_palette

PlotUtils.tritanopic

PlotUtils.weighted_color_mean

PlotUtils.whitebalance

PlotUtils.xyY

PlotUtils.xyYA