Docstrings

Juno.@enter — Macro.

@enter ex

Step into the function call in ex.

Juno.@profiler — Macro.

@profiler

Clear currently collected profile traces, profile the provided expression and show it via Juno.profiler().

Juno.@progress — Macro.

@progress [name="", threshold=0.005] for i = ..., j = ..., ...
@progress [name="", threshold=0.005] x = [... for i = ..., j = ..., ...]

Show a progress meter named name for the given loop or array comprehension if possible. Update frequency is limited by threshold (one update per 0.5% of progress by default).

Juno.@render

Juno.@sh

Juno.Clipboard

Juno.Console

Juno.Copyable

Juno.Editor

Juno.Inline

Juno.Juno

Juno.LazyTree

Juno.Link

Juno.Media

Juno.Model

Juno.PROGRESSLEVEL

Juno.PlotPane

Juno.Row

Juno.SubTree

Juno.Table

Juno.Tree

Juno.UNDEF

Juno._active

Juno._comprehension

Juno._for

Juno._progress

Juno.activate

Juno.breakpoint

Juno.clearconsole — Function.

clearconsole()

Clear the console if Juno is used; does nothing otherwise.

Juno.connect

Juno.defaultrepr — Function.

defaultrepr(x, lazy = false)

render fallback for types without any specialized show methods.

If lazy is true, then the type's fields will be loaded lazily when expanding the tree. This is useful when the fields contain big elements that might need to be inspectable.

Can be used by packages to restore Juno's default printing if they have defined a show method that should not be used by Juno:

Juno.render(i::Juno.Inline, x::myType) = Juno.render(i, Juno.defaultrepr(x))

Juno.dims

Juno.errmsg

Juno.errtrace

Juno.eval

Juno.fade

Juno.icon

Juno.include

Juno.info — Function.

info(msg)

Show the given message in Juno's console using blue styling, or fall back to Base.info.

In a package, you can use import Juno: info to replace the default version with this one.

Juno.input — Function.

input(prompt = "") -> "..."

Prompt the user to input some text, and return it. Optionally display a prompt.

Juno.interleave

Juno.isactive — Function.

isactive()

Will return true when the current Julia process is connected to a running Juno frontend.

Juno.isprecompiling

Juno.limit

Juno.link

Juno.media

Juno.notify — Function.

notify(msg)

Display msg as an OS specific notification.

Useful for signaling the end of a long running computation or similar. This disregards the Notifications setting in julia-client. Falls back to info(msg) in other environments.

Juno.plotsize — Function.

plotsize()

Get the size of Juno's plot pane in px. Returns [100, 100] as a fallback value.

Juno.profiler — Function.

profiler()

Show currently collected profile information as an in-editor flamechart.

Juno.profiletree — Function.

profiletree()

Show currently collected profile information in tree-form. Falls back to Profile.print().

Juno.progress — Function.

progress(f::Function; name = "", msg = "")

Evaluates f with id as its argument and makes sure to destroy the progress bar afterwards. To update the progress bar in f you can call a logging statement like @info or even just @logmsg with _id=id and progress as arguments.

progress can take either of the following values:

0 <= progress < 1: create or update progress bar

progress == nothing || progress = NaN: set progress bar to indeterminate progress

progress > 1 || progress == "done": destroy progress bar

The logging message (e.g. "foo" in @info "foo") will be used as the progress bar's name.

Juno.progress() do id
    for i = 1:10
        sleep(0.5)
        @info "iterating" progress=i/10 _id=id
    end
end

Juno.selector — Function.

selector([xs...]) -> x

Allow the user to select one of the xs.

xs should be an iterator of strings. Currently there is no fallback in other environments.

Juno.setactive!

Juno.structure — Function.

structure(x)

Display x's underlying representation, rather than using its normal display method.

For example, structure(:(2x+1)) displays the Expr object with its head and args fields instead of printing the expression.

Juno.syntaxcolors — Function.

syntaxcolors(selectors = Atom.SELECTORS)::Dict{String, UInt32}

Get the colors used by the current Atom theme. selectors should be a Dict{String, Vector{String}} which assigns a css selector (e.g. syntax--julia) to a name (e.g. variable).

Juno.undefs

Juno.view