Docstrings

CUDAapi.@trace

CUDAapi.CUDAapi

CUDAapi.MAJOR_VERSION

CUDAapi.MINOR_VERSION

CUDAapi.PATCH_LEVEL

CUDAapi.Toolchain

CUDAapi.VersionRange

CUDAapi.cuda_gcc_db

CUDAapi.cuda_msvc_db

CUDAapi.cuda_names

CUDAapi.cuda_versions

CUDAapi.dev_cuda_db

CUDAapi.dev_llvm_db

CUDAapi.devices_for_cuda

CUDAapi.devices_for_llvm

CUDAapi.eval

CUDAapi.find_binary — Function.

find_binary(names; locations=String[])

Similar to find_library, performs an exhaustive search for a binary in various subdirectories of locations, and finally PATH.

CUDAapi.find_cuda_binary

CUDAapi.find_cuda_library

CUDAapi.find_host_compiler

CUDAapi.find_libdevice — Function.

find_libdevice(targets::Vector{VersionNumber}, toolkit_dirs::Vector{String})

Look for the CUDA device library supporting targets in any of the CUDA toolkit directories toolkit_dirs. On CUDA >= 9.0, a single library unified library is discovered and returned as a string. On older toolkits, individual libraries for each of the targets are returned as a vector of strings.

CUDAapi.find_library — Function.

find_library(names; locations=String[], versions=VersionNumber[], word_size=Sys.WORD_SIZE)

Wrapper for Libdl.find_library, performing a more exhaustive search:

variants of the library name (including version numbers, platform-specific tags, etc);

various subdirectories of the locations list, and finally system library directories.

Returns the full path to the library.

CUDAapi.find_toolchain

CUDAapi.find_toolkit — Function.

find_toolkit()::Vector{String}

Look for directories where (parts of) the CUDA toolkit might be installed. This returns a (possibly empty) list of paths that can be used as an argument to other discovery functions.

The behavior of this function can be overridden by defining the CUDA_PATH, CUDA_HOME or CUDA_ROOT environment variables, which should point to the root of the CUDA toolkit.

CUDAapi.find_toolkit_version

CUDAapi.gcc_supported

CUDAapi.highest

CUDAapi.include

CUDAapi.isa_cuda_db

CUDAapi.isa_llvm_db

CUDAapi.isas_for_cuda

CUDAapi.isas_for_llvm

CUDAapi.libraryPropertyType

CUDAapi.lowest

CUDAapi.msvc_supported

CUDAapi.resolve

CUDAapi.shader

CUDAapi.strip_minor

CUDAapi.strip_patch

CUDAapi.valid_dirs