Docstrings

LCMCore.@checked_lib

LCMCore.@lcmtypesetup — Macro.

@lcmtypesetup(lcmtype, dimensioninfos...)

Generate the following methods for a concrete LCMType subtype (say MyType):

dimensions(x::MyType, ::Val{fieldsym}), for all fields

fingerprint(::Type{MyType})

The lcmtype argument should be the name of a concrete LCMType subtype. The dimensioninfos arguments can be used to define which fields determine the size of which variable-size array fields. Each dimensioninfos element should have the form

arrayfieldname => (size1, size2, ...)

where arrayfieldname is the name of an Array field and size1, size2 etc. are the dimensions of the Array, where both integers and field names that come before arrayfieldname in the type definition may be used.

Examples

mutable struct MyType <: LCMType
    alength::Int32
    c_inner_length::Int32
    a::Vector{Float64}
    b::SVector{3, Float32}
    c::Matrix{Int64}
end

@lcmtypesetup(MyType,
    a => (alength, ),
    c => (3, c_inner_length)
)

LCMCore.DimensionMode

LCMCore.FingerprintException

LCMCore.LCM

LCMCore.LCMCore

LCMCore.LCMDimension

LCMCore.LCMDimensionConst

LCMCore.LCMDimensionVar

LCMCore.LCMLog

LCMCore.LCMPrimitive

LCMCore.LCMType — Type.

LCMType

Supertype of concrete Julia structs that represent LCM message types.

Subtypes must be mutable structs and may use the following field types:

Bool

numeric types: Int8, Int16, Int32, Int64, Float32, Float64

bytes (encoded in the same was as Int8): UInt8;

String;

another LCMType;

Vector or a subtype of StaticVector, for which the element type must also be

one of the previously specified types or another Vector or StaticVector.

Any size fields must come before the Vector fields to which they correspond.

The following methods must be defined for a concrete subtype of LCMType (say MyType):

dimensions(x::MyType, ::Val{fieldsym})

fingerprint(::Type{MyType})

Note that the @lcmtypesetup macro can be used to generate these methods automatically.

Also note that ideally, all of these methods would be generated from the LCM message type definition, but that is currently not the case.

LCMCore.LCM_CONST

LCMCore.LCM_VAR

LCMCore.NetworkByteOrderEncoded

LCMCore.RecvBuf

LCMCore.ReverseDimIndices

LCMCore.Subscription

LCMCore.SubscriptionOptions

LCMCore._decodefield

LCMCore._defaultval

LCMCore.basehash

LCMCore.check_channel_name

LCMCore.check_loopback_multicast

LCMCore.check_multicast_routing

LCMCore.check_valid — Function.

check_valid(x::LCMType)

Check that the array sizes of x match their corresponding size fields.

LCMCore.checkfingerprint

LCMCore.close

LCMCore.computehash

LCMCore.decode

LCMCore.decode!

LCMCore.decode_in_place — Function.

decode_in_place(T)

Specify whether type T should be decoded in place, i.e. whether to use a decodefield! method instead of a decodefield method.

LCMCore.decodefield

LCMCore.decodefield!

LCMCore.defaultval

LCMCore.depsjl

LCMCore.dimensionmode

LCMCore.dimensions — Function.

dimensions(x::LCMType, ::Val{fieldsym})

Return a tuple of LCMDimensions describing the size of getfield(x, fieldsym).

LCMCore.encode — Function.

encode(io::IO, x::LCMType)

Write an LCM byte representation of x to io.

LCMCore.encodefield

LCMCore.eval

LCMCore.evaldims

LCMCore.filedescriptor

LCMCore.fingerprint — Function.

fingerprint(::Type{T}) where T<:LCMType

Return the fingerprint of LCM type T as an Int64 (exactly 8 bytes).

LCMCore.get_queue_size

LCMCore.handle

LCMCore.hashupdate

LCMCore.include

LCMCore.isgood

LCMCore.isprimitive

LCMCore.lcm_eventlog_create

LCMCore.lcm_eventlog_destroy

LCMCore.lcm_eventlog_event_t

LCMCore.lcm_eventlog_read_next_event

LCMCore.lcm_eventlog_t

LCMCore.lcm_handle

LCMCore.lcm_prefix

LCMCore.lcmtypename

LCMCore.liblcm

LCMCore.loopback_interface

LCMCore.loopback_multicast_setup_advice

LCMCore.makedim

LCMCore.onresponse

LCMCore.publish

LCMCore.read_next_event

LCMCore.resizearrayfield!

LCMCore.reversedimindices

LCMCore.reversedims

LCMCore.set_queue_capacity

LCMCore.sign_extended_right_shift

LCMCore.sizestring

LCMCore.subscribe

LCMCore.troubleshoot

LCMCore.unsubscribe