Docstrings

QuartzImageIO.CFBooleanGetValue

QuartzImageIO.CFDataCreate

QuartzImageIO.CFDataGetBytePtr

QuartzImageIO.CFDictionaryGetValue

QuartzImageIO.CFDictionaryRef

QuartzImageIO.CFNumberGetValue

QuartzImageIO.CFRelease

QuartzImageIO.CFShow

QuartzImageIO.CFStringGetCString

QuartzImageIO.CFStringRef

QuartzImageIO.CFURLCreateWithFileSystemPath

QuartzImageIO.CFURLRef

QuartzImageIO.CGBitmapContextCreate

QuartzImageIO.CGBitmapContextCreateImage

QuartzImageIO.CGColorSpaceCreateDeviceRGB

QuartzImageIO.CGColorSpaceCreateWithName

QuartzImageIO.CGColorSpaceRef

QuartzImageIO.CGContextRef

QuartzImageIO.CGDataProviderCopyData

QuartzImageIO.CGDisplayNoErr

QuartzImageIO.CGEventNoErr

QuartzImageIO.CGFLOAT_DEFINED

QuartzImageIO.CGFLOAT_IS_DOUBLE

QuartzImageIO.CGFLOAT_MAX

QuartzImageIO.CGFLOAT_MIN

QuartzImageIO.CGGlyphMax

QuartzImageIO.CGGlyphMin

QuartzImageIO.CGImageDestinationAddImage

QuartzImageIO.CGImageDestinationCreateWithURL

QuartzImageIO.CGImageDestinationFinalize

QuartzImageIO.CGImageDestinationRef

QuartzImageIO.CGImageGetAlphaInfo

QuartzImageIO.CGImageGetBitsPerComponent

QuartzImageIO.CGImageGetBitsPerPixel

QuartzImageIO.CGImageGetColorSpace

QuartzImageIO.CGImageGetDataProvider

QuartzImageIO.CGImageRef

QuartzImageIO.CGImageRelease

QuartzImageIO.CGImageSourceCopyPropertiesAtIndex

QuartzImageIO.CGImageSourceCreateImageAtIndex

QuartzImageIO.CGImageSourceCreateWithData

QuartzImageIO.CGImageSourceCreateWithURL

QuartzImageIO.CGImageSourceGetCount

QuartzImageIO.CGImageSourceGetStatus

QuartzImageIO.CGImageSourceGetStatusAtIndex

QuartzImageIO.CGImageSourceGetType

QuartzImageIO.CGPDFDataFormatJPEG2000

QuartzImageIO.CGPDFDataFormatJPEGEncoded

QuartzImageIO.CGPDFDataFormatRaw

QuartzImageIO.CGRectMaxXEdge

QuartzImageIO.CGRectMaxYEdge

QuartzImageIO.CGRectMinXEdge

QuartzImageIO.CGRectMinYEdge

QuartzImageIO.CGVECTOR_DEFINED

QuartzImageIO.Color1

QuartzImageIO.Color2

QuartzImageIO.Color3

QuartzImageIO.Color4

QuartzImageIO.CopyImagePixels

QuartzImageIO.NSString

QuartzImageIO.QuartzImageIO

QuartzImageIO.WritingImageFailed

QuartzImageIO.__init__

QuartzImageIO.alpha_and_depth

QuartzImageIO.check_null — Function.

check_null(x)

Triggers an error if x is NULL, else returns x

QuartzImageIO.eval

QuartzImageIO.fillcolor!

QuartzImageIO.fillgray!

QuartzImageIO.fillgrayalpha!

QuartzImageIO.foundation

QuartzImageIO.getblob

QuartzImageIO.imageio

QuartzImageIO.include

QuartzImageIO.kCFNumberCFIndexType

QuartzImageIO.kCFNumberCGFloatType

QuartzImageIO.kCFNumberCharType

QuartzImageIO.kCFNumberDoubleType

QuartzImageIO.kCFNumberFloat32Type

QuartzImageIO.kCFNumberFloat64Type

QuartzImageIO.kCFNumberFloatType

QuartzImageIO.kCFNumberIntType

QuartzImageIO.kCFNumberLongLongType

QuartzImageIO.kCFNumberLongType

QuartzImageIO.kCFNumberMaxType

QuartzImageIO.kCFNumberNSIntegerType

QuartzImageIO.kCFNumberSInt16Type

QuartzImageIO.kCFNumberSInt32Type

QuartzImageIO.kCFNumberSInt64Type

QuartzImageIO.kCFNumberSInt8Type

QuartzImageIO.kCFNumberShortType

QuartzImageIO.kCGAnnotatedSessionEventTap

QuartzImageIO.kCGAnyInputEventType

QuartzImageIO.kCGAssistiveTechHighWindowLevel

QuartzImageIO.kCGAssistiveTechHighWindowLevelKey

QuartzImageIO.kCGBackingStoreBuffered

QuartzImageIO.kCGBackingStoreNonretained

QuartzImageIO.kCGBackingStoreRetained

QuartzImageIO.kCGBackstopMenuLevel

QuartzImageIO.kCGBackstopMenuLevelKey

QuartzImageIO.kCGBaseWindowLevel

QuartzImageIO.kCGBaseWindowLevelKey

QuartzImageIO.kCGBitmapAlphaInfoMask

QuartzImageIO.kCGBitmapByteOrder16Big

QuartzImageIO.kCGBitmapByteOrder16Host

QuartzImageIO.kCGBitmapByteOrder16Little

QuartzImageIO.kCGBitmapByteOrder32Big

QuartzImageIO.kCGBitmapByteOrder32Host

QuartzImageIO.kCGBitmapByteOrder32Little

QuartzImageIO.kCGBitmapByteOrderDefault

QuartzImageIO.kCGBitmapByteOrderMask

QuartzImageIO.kCGBitmapFloatComponents

QuartzImageIO.kCGBitmapFloatInfoMask

QuartzImageIO.kCGBlendModeClear

QuartzImageIO.kCGBlendModeColor

QuartzImageIO.kCGBlendModeColorBurn

QuartzImageIO.kCGBlendModeColorDodge

QuartzImageIO.kCGBlendModeCopy

QuartzImageIO.kCGBlendModeDarken

QuartzImageIO.kCGBlendModeDestinationAtop

QuartzImageIO.kCGBlendModeDestinationIn

QuartzImageIO.kCGBlendModeDestinationOut

QuartzImageIO.kCGBlendModeDestinationOver

QuartzImageIO.kCGBlendModeDifference

QuartzImageIO.kCGBlendModeExclusion

QuartzImageIO.kCGBlendModeHardLight

QuartzImageIO.kCGBlendModeHue

QuartzImageIO.kCGBlendModeLighten

QuartzImageIO.kCGBlendModeLuminosity

QuartzImageIO.kCGBlendModeMultiply

QuartzImageIO.kCGBlendModeNormal

QuartzImageIO.kCGBlendModeOverlay

QuartzImageIO.kCGBlendModePlusDarker

QuartzImageIO.kCGBlendModePlusLighter

QuartzImageIO.kCGBlendModeSaturation

QuartzImageIO.kCGBlendModeScreen

QuartzImageIO.kCGBlendModeSoftLight

QuartzImageIO.kCGBlendModeSourceAtop

QuartzImageIO.kCGBlendModeSourceIn

QuartzImageIO.kCGBlendModeSourceOut

QuartzImageIO.kCGBlendModeXOR

QuartzImageIO.kCGCaptureNoFill

QuartzImageIO.kCGCaptureNoOptions

QuartzImageIO.kCGColorConversionTransformApplySpace

QuartzImageIO.kCGColorConversionTransformFromSpace

QuartzImageIO.kCGColorConversionTransformToSpace

QuartzImageIO.kCGColorSpaceModelCMYK

QuartzImageIO.kCGColorSpaceModelDeviceN

QuartzImageIO.kCGColorSpaceModelIndexed

QuartzImageIO.kCGColorSpaceModelLab

QuartzImageIO.kCGColorSpaceModelMonochrome

QuartzImageIO.kCGColorSpaceModelPattern

QuartzImageIO.kCGColorSpaceModelRGB

QuartzImageIO.kCGColorSpaceModelUnknown

QuartzImageIO.kCGConfigureForAppOnly

QuartzImageIO.kCGConfigureForSession

QuartzImageIO.kCGConfigurePermanently

QuartzImageIO.kCGCursorWindowLevel

QuartzImageIO.kCGCursorWindowLevelKey

QuartzImageIO.kCGDesktopIconWindowLevel

QuartzImageIO.kCGDesktopIconWindowLevelKey

QuartzImageIO.kCGDesktopWindowLevel

QuartzImageIO.kCGDesktopWindowLevelKey

QuartzImageIO.kCGDirectMainDisplay

QuartzImageIO.kCGDisplayAddFlag

QuartzImageIO.kCGDisplayBeginConfigurationFlag

QuartzImageIO.kCGDisplayBlendNormal

QuartzImageIO.kCGDisplayBlendSolidColor

QuartzImageIO.kCGDisplayDesktopShapeChangedFlag

QuartzImageIO.kCGDisplayDisabledFlag

QuartzImageIO.kCGDisplayEnabledFlag

QuartzImageIO.kCGDisplayFadeReservationInvalidToken

QuartzImageIO.kCGDisplayMirrorFlag

QuartzImageIO.kCGDisplayMovedFlag

QuartzImageIO.kCGDisplayRemoveFlag

QuartzImageIO.kCGDisplaySetMainFlag

QuartzImageIO.kCGDisplaySetModeFlag

QuartzImageIO.kCGDisplayStreamFrameStatusFrameBlank

QuartzImageIO.kCGDisplayStreamFrameStatusFrameComplete

QuartzImageIO.kCGDisplayStreamFrameStatusFrameIdle

QuartzImageIO.kCGDisplayStreamFrameStatusStopped

QuartzImageIO.kCGDisplayStreamUpdateDirtyRects

QuartzImageIO.kCGDisplayStreamUpdateMovedRects

QuartzImageIO.kCGDisplayStreamUpdateReducedDirtyRects

QuartzImageIO.kCGDisplayStreamUpdateRefreshedRects

QuartzImageIO.kCGDisplayUnMirrorFlag

QuartzImageIO.kCGDockWindowLevel

QuartzImageIO.kCGDockWindowLevelKey

QuartzImageIO.kCGDraggingWindowLevel

QuartzImageIO.kCGDraggingWindowLevelKey

QuartzImageIO.kCGEncodingFontSpecific

QuartzImageIO.kCGEncodingMacRoman

QuartzImageIO.kCGErrorCannotComplete

QuartzImageIO.kCGErrorFailure

QuartzImageIO.kCGErrorIllegalArgument

QuartzImageIO.kCGErrorInvalidConnection

QuartzImageIO.kCGErrorInvalidContext

QuartzImageIO.kCGErrorInvalidOperation

QuartzImageIO.kCGErrorNoneAvailable

QuartzImageIO.kCGErrorNotImplemented

QuartzImageIO.kCGErrorRangeCheck

QuartzImageIO.kCGErrorSuccess

QuartzImageIO.kCGErrorTypeCheck

QuartzImageIO.kCGEventFilterMaskPermitAllEvents

QuartzImageIO.kCGEventFilterMaskPermitLocalKeyboardEvents

QuartzImageIO.kCGEventFilterMaskPermitLocalMouseEvents

QuartzImageIO.kCGEventFilterMaskPermitSystemDefinedEvents

QuartzImageIO.kCGEventFlagMaskAlphaShift

QuartzImageIO.kCGEventFlagMaskAlternate

QuartzImageIO.kCGEventFlagMaskCommand

QuartzImageIO.kCGEventFlagMaskControl

QuartzImageIO.kCGEventFlagMaskHelp

QuartzImageIO.kCGEventFlagMaskNonCoalesced

QuartzImageIO.kCGEventFlagMaskNumericPad

QuartzImageIO.kCGEventFlagMaskSecondaryFn

QuartzImageIO.kCGEventFlagMaskShift

QuartzImageIO.kCGEventFlagsChanged

QuartzImageIO.kCGEventKeyDown

QuartzImageIO.kCGEventKeyUp

QuartzImageIO.kCGEventLeftMouseDown

QuartzImageIO.kCGEventLeftMouseDragged

QuartzImageIO.kCGEventLeftMouseUp

QuartzImageIO.kCGEventMaskForAllEvents

QuartzImageIO.kCGEventMouseMoved

QuartzImageIO.kCGEventMouseSubtypeDefault

QuartzImageIO.kCGEventMouseSubtypeTabletPoint

QuartzImageIO.kCGEventMouseSubtypeTabletProximity

QuartzImageIO.kCGEventNull

QuartzImageIO.kCGEventOtherMouseDown

QuartzImageIO.kCGEventOtherMouseDragged

QuartzImageIO.kCGEventOtherMouseUp

QuartzImageIO.kCGEventRightMouseDown

QuartzImageIO.kCGEventRightMouseDragged

QuartzImageIO.kCGEventRightMouseUp

QuartzImageIO.kCGEventScrollWheel

QuartzImageIO.kCGEventSourceGroupID

QuartzImageIO.kCGEventSourceStateCombinedSessionState

QuartzImageIO.kCGEventSourceStateHIDSystemState

QuartzImageIO.kCGEventSourceStateID

QuartzImageIO.kCGEventSourceStatePrivate

QuartzImageIO.kCGEventSourceUnixProcessID

QuartzImageIO.kCGEventSourceUserData

QuartzImageIO.kCGEventSourceUserID

QuartzImageIO.kCGEventSuppressionStateRemoteMouseDrag

QuartzImageIO.kCGEventSuppressionStateSuppressionInterval

QuartzImageIO.kCGEventSupressionStateRemoteMouseDrag

QuartzImageIO.kCGEventSupressionStateSupressionInterval

QuartzImageIO.kCGEventTabletPointer

QuartzImageIO.kCGEventTabletProximity

QuartzImageIO.kCGEventTapDisabledByTimeout

QuartzImageIO.kCGEventTapDisabledByUserInput

QuartzImageIO.kCGEventTapOptionDefault

QuartzImageIO.kCGEventTapOptionListenOnly

QuartzImageIO.kCGEventTargetProcessSerialNumber

QuartzImageIO.kCGEventTargetUnixProcessID

QuartzImageIO.kCGFloatingWindowLevel

QuartzImageIO.kCGFloatingWindowLevelKey

QuartzImageIO.kCGFontIndexInvalid

QuartzImageIO.kCGFontIndexMax

QuartzImageIO.kCGFontPostScriptFormatType1

QuartzImageIO.kCGFontPostScriptFormatType3

QuartzImageIO.kCGFontPostScriptFormatType42

QuartzImageIO.kCGGesturePhaseBegan

QuartzImageIO.kCGGesturePhaseCancelled

QuartzImageIO.kCGGesturePhaseChanged

QuartzImageIO.kCGGesturePhaseEnded

QuartzImageIO.kCGGesturePhaseMayBegin

QuartzImageIO.kCGGesturePhaseNone

QuartzImageIO.kCGGlyphMax

QuartzImageIO.kCGGradientDrawsAfterEndLocation

QuartzImageIO.kCGGradientDrawsBeforeStartLocation

QuartzImageIO.kCGHIDEventTap

QuartzImageIO.kCGHeadInsertEventTap

QuartzImageIO.kCGHelpWindowLevel

QuartzImageIO.kCGHelpWindowLevelKey

QuartzImageIO.kCGImageAlphaFirst

QuartzImageIO.kCGImageAlphaLast

QuartzImageIO.kCGImageAlphaNone

QuartzImageIO.kCGImageAlphaNoneSkipFirst

QuartzImageIO.kCGImageAlphaNoneSkipLast

QuartzImageIO.kCGImageAlphaOnly

QuartzImageIO.kCGImageAlphaPremultipliedFirst

QuartzImageIO.kCGImageAlphaPremultipliedLast

QuartzImageIO.kCGImageByteOrder16Big

QuartzImageIO.kCGImageByteOrder16Little

QuartzImageIO.kCGImageByteOrder32Big

QuartzImageIO.kCGImageByteOrder32Little

QuartzImageIO.kCGImageByteOrderMask

QuartzImageIO.kCGInterpolationDefault

QuartzImageIO.kCGInterpolationHigh

QuartzImageIO.kCGInterpolationLow

QuartzImageIO.kCGInterpolationMedium

QuartzImageIO.kCGInterpolationNone

QuartzImageIO.kCGKeyboardEventAutorepeat

QuartzImageIO.kCGKeyboardEventKeyboardType

QuartzImageIO.kCGKeyboardEventKeycode

QuartzImageIO.kCGLineCapButt

QuartzImageIO.kCGLineCapRound

QuartzImageIO.kCGLineCapSquare

QuartzImageIO.kCGLineJoinBevel

QuartzImageIO.kCGLineJoinMiter

QuartzImageIO.kCGLineJoinRound

QuartzImageIO.kCGMainMenuWindowLevel

QuartzImageIO.kCGMainMenuWindowLevelKey

QuartzImageIO.kCGMaxDisplayReservationInterval

QuartzImageIO.kCGMaximumWindowLevel

QuartzImageIO.kCGMaximumWindowLevelKey

QuartzImageIO.kCGMinimumWindowLevel

QuartzImageIO.kCGMinimumWindowLevelKey

QuartzImageIO.kCGModalPanelWindowLevel

QuartzImageIO.kCGModalPanelWindowLevelKey

QuartzImageIO.kCGMomentumScrollPhaseBegin

QuartzImageIO.kCGMomentumScrollPhaseContinue

QuartzImageIO.kCGMomentumScrollPhaseEnd

QuartzImageIO.kCGMomentumScrollPhaseNone

QuartzImageIO.kCGMouseButtonCenter

QuartzImageIO.kCGMouseButtonLeft

QuartzImageIO.kCGMouseButtonRight

QuartzImageIO.kCGMouseDownEventMaskingDeadSwitchTimeout

QuartzImageIO.kCGMouseEventButtonNumber

QuartzImageIO.kCGMouseEventClickState

QuartzImageIO.kCGMouseEventDeltaX

QuartzImageIO.kCGMouseEventDeltaY

QuartzImageIO.kCGMouseEventInstantMouser

QuartzImageIO.kCGMouseEventNumber

QuartzImageIO.kCGMouseEventPressure

QuartzImageIO.kCGMouseEventSubtype

QuartzImageIO.kCGMouseEventWindowUnderMousePointer

QuartzImageIO.kCGMouseEventWindowUnderMousePointerThatCanHandleThisEvent

QuartzImageIO.kCGNormalWindowLevel

QuartzImageIO.kCGNormalWindowLevelKey

QuartzImageIO.kCGNullDirectDisplay

QuartzImageIO.kCGNullWindowID

QuartzImageIO.kCGNumReservedWindowLevels

QuartzImageIO.kCGNumberOfEventSuppressionStates

QuartzImageIO.kCGNumberOfEventSupressionStates

QuartzImageIO.kCGNumberOfWindowLevelKeys

QuartzImageIO.kCGOverlayWindowLevel

QuartzImageIO.kCGOverlayWindowLevelKey

QuartzImageIO.kCGPDFArtBox

QuartzImageIO.kCGPDFBleedBox

QuartzImageIO.kCGPDFCropBox

QuartzImageIO.kCGPDFMediaBox

QuartzImageIO.kCGPDFObjectTypeArray

QuartzImageIO.kCGPDFObjectTypeBoolean

QuartzImageIO.kCGPDFObjectTypeDictionary

QuartzImageIO.kCGPDFObjectTypeInteger

QuartzImageIO.kCGPDFObjectTypeName

QuartzImageIO.kCGPDFObjectTypeNull

QuartzImageIO.kCGPDFObjectTypeReal

QuartzImageIO.kCGPDFObjectTypeStream

QuartzImageIO.kCGPDFObjectTypeString

QuartzImageIO.kCGPDFTrimBox

QuartzImageIO.kCGPathEOFill

QuartzImageIO.kCGPathEOFillStroke

QuartzImageIO.kCGPathElementAddCurveToPoint

QuartzImageIO.kCGPathElementAddLineToPoint

QuartzImageIO.kCGPathElementAddQuadCurveToPoint

QuartzImageIO.kCGPathElementCloseSubpath

QuartzImageIO.kCGPathElementMoveToPoint

QuartzImageIO.kCGPathFill

QuartzImageIO.kCGPathFillStroke

QuartzImageIO.kCGPathStroke

QuartzImageIO.kCGPatternTilingConstantSpacing

QuartzImageIO.kCGPatternTilingConstantSpacingMinimalDistortion

QuartzImageIO.kCGPatternTilingNoDistortion

QuartzImageIO.kCGPopUpMenuWindowLevel

QuartzImageIO.kCGPopUpMenuWindowLevelKey

QuartzImageIO.kCGRenderingIntentAbsoluteColorimetric

QuartzImageIO.kCGRenderingIntentDefault

QuartzImageIO.kCGRenderingIntentPerceptual

QuartzImageIO.kCGRenderingIntentRelativeColorimetric

QuartzImageIO.kCGRenderingIntentSaturation

QuartzImageIO.kCGScreenSaverWindowLevel

QuartzImageIO.kCGScreenSaverWindowLevelKey

QuartzImageIO.kCGScreenUpdateOperationMove

QuartzImageIO.kCGScreenUpdateOperationReducedDirtyRectangleCount

QuartzImageIO.kCGScreenUpdateOperationRefresh

QuartzImageIO.kCGScrollEventUnitLine

QuartzImageIO.kCGScrollEventUnitPixel

QuartzImageIO.kCGScrollPhaseBegan

QuartzImageIO.kCGScrollPhaseCancelled

QuartzImageIO.kCGScrollPhaseChanged

QuartzImageIO.kCGScrollPhaseEnded

QuartzImageIO.kCGScrollPhaseMayBegin

QuartzImageIO.kCGScrollWheelEventDeltaAxis1

QuartzImageIO.kCGScrollWheelEventDeltaAxis2

QuartzImageIO.kCGScrollWheelEventDeltaAxis3

QuartzImageIO.kCGScrollWheelEventFixedPtDeltaAxis1

QuartzImageIO.kCGScrollWheelEventFixedPtDeltaAxis2

QuartzImageIO.kCGScrollWheelEventFixedPtDeltaAxis3

QuartzImageIO.kCGScrollWheelEventInstantMouser

QuartzImageIO.kCGScrollWheelEventIsContinuous

QuartzImageIO.kCGScrollWheelEventMomentumPhase

QuartzImageIO.kCGScrollWheelEventPointDeltaAxis1

QuartzImageIO.kCGScrollWheelEventPointDeltaAxis2

QuartzImageIO.kCGScrollWheelEventPointDeltaAxis3

QuartzImageIO.kCGScrollWheelEventScrollCount

QuartzImageIO.kCGScrollWheelEventScrollPhase

QuartzImageIO.kCGSessionEventTap

QuartzImageIO.kCGStatusWindowLevel

QuartzImageIO.kCGStatusWindowLevelKey

QuartzImageIO.kCGTabletEventDeviceID

QuartzImageIO.kCGTabletEventPointButtons

QuartzImageIO.kCGTabletEventPointPressure

QuartzImageIO.kCGTabletEventPointX

QuartzImageIO.kCGTabletEventPointY

QuartzImageIO.kCGTabletEventPointZ

QuartzImageIO.kCGTabletEventRotation

QuartzImageIO.kCGTabletEventTangentialPressure

QuartzImageIO.kCGTabletEventTiltX

QuartzImageIO.kCGTabletEventTiltY

QuartzImageIO.kCGTabletEventVendor1

QuartzImageIO.kCGTabletEventVendor2

QuartzImageIO.kCGTabletEventVendor3

QuartzImageIO.kCGTabletProximityEventCapabilityMask

QuartzImageIO.kCGTabletProximityEventDeviceID

QuartzImageIO.kCGTabletProximityEventEnterProximity

QuartzImageIO.kCGTabletProximityEventPointerID

QuartzImageIO.kCGTabletProximityEventPointerType

QuartzImageIO.kCGTabletProximityEventSystemTabletID

QuartzImageIO.kCGTabletProximityEventTabletID

QuartzImageIO.kCGTabletProximityEventVendorID

QuartzImageIO.kCGTabletProximityEventVendorPointerSerialNumber

QuartzImageIO.kCGTabletProximityEventVendorPointerType

QuartzImageIO.kCGTabletProximityEventVendorUniqueID

QuartzImageIO.kCGTailAppendEventTap

QuartzImageIO.kCGTextClip

QuartzImageIO.kCGTextFill

QuartzImageIO.kCGTextFillClip

QuartzImageIO.kCGTextFillStroke

QuartzImageIO.kCGTextFillStrokeClip

QuartzImageIO.kCGTextInvisible

QuartzImageIO.kCGTextStroke

QuartzImageIO.kCGTextStrokeClip

QuartzImageIO.kCGTornOffMenuWindowLevel

QuartzImageIO.kCGTornOffMenuWindowLevelKey

QuartzImageIO.kCGUtilityWindowLevel

QuartzImageIO.kCGUtilityWindowLevelKey

QuartzImageIO.kCGWindowBackingCFNumberType

QuartzImageIO.kCGWindowIDCFNumberType

QuartzImageIO.kCGWindowImageBestResolution

QuartzImageIO.kCGWindowImageBoundsIgnoreFraming

QuartzImageIO.kCGWindowImageDefault

QuartzImageIO.kCGWindowImageNominalResolution

QuartzImageIO.kCGWindowImageOnlyShadows

QuartzImageIO.kCGWindowImageShouldBeOpaque

QuartzImageIO.kCGWindowListExcludeDesktopElements

QuartzImageIO.kCGWindowListOptionAll

QuartzImageIO.kCGWindowListOptionIncludingWindow

QuartzImageIO.kCGWindowListOptionOnScreenAboveWindow

QuartzImageIO.kCGWindowListOptionOnScreenBelowWindow

QuartzImageIO.kCGWindowListOptionOnScreenOnly

QuartzImageIO.kCGWindowSharingCFNumberType

QuartzImageIO.kCGWindowSharingNone

QuartzImageIO.kCGWindowSharingReadOnly

QuartzImageIO.kCGWindowSharingReadWrite

QuartzImageIO.load

QuartzImageIO.load_

QuartzImageIO.mapCG

QuartzImageIO.ogc

QuartzImageIO.oms

QuartzImageIO.permutedims_horizontal

QuartzImageIO.read_and_release_imgsrc

QuartzImageIO.save

QuartzImageIO.save_ — Function.

save_(f, img, image_type)

f is the file to save to, of type FileIO.File{DataFormat}

image_type should be one of Apple's image types (eg. "public.jpeg")

permute_horizontal, if true, will transpose the image (flip x and y)

mapi is the mapping to apply to the data before saving. Defaults to identity. A useful alternative value is clamp01nan.

QuartzImageIO.selector

QuartzImageIO.to_contiguous

QuartzImageIO.to_explicit