Docstrings

MathOptInterfaceMosek.AffineFunction

MathOptInterfaceMosek.ConstraintMap

MathOptInterfaceMosek.DEBUG

MathOptInterfaceMosek.LinearDomain

MathOptInterfaceMosek.LinearFunction

MathOptInterfaceMosek.LinkedInts

MathOptInterfaceMosek.MOI

MathOptInterfaceMosek.MOIU

MathOptInterfaceMosek.MathOptInterfaceMosek

MathOptInterfaceMosek.MosekModel — Type.

MosekModel <: MathOptInterface.AbstractModel

Linear variables and constraint can be deleted. For some reason MOSEK does not support deleting PSD variables.

Note also that adding variables and constraints will permanently add some (currently between 1 and 3) Int64s that a delete! will not remove. This ensures that Indices (Variable and constraint) that are deleted are thereafter invalid.

MathOptInterfaceMosek.MosekOptimizer

MathOptInterfaceMosek.MosekSolution

MathOptInterfaceMosek.ObjF

MathOptInterfaceMosek.PositiveSemidefiniteCone

MathOptInterfaceMosek.ScalarIntegerDomain

MathOptInterfaceMosek.ScalarLinearDomain

MathOptInterfaceMosek.VariableFunction

MathOptInterfaceMosek.VectorCone

MathOptInterfaceMosek.VectorLinearDomain

MathOptInterfaceMosek._sympackedto

MathOptInterfaceMosek.abstractset2ct

MathOptInterfaceMosek.addbound!

MathOptInterfaceMosek.addlhsblock!

MathOptInterfaceMosek.addvarconstr

MathOptInterfaceMosek.allocateconstraints

MathOptInterfaceMosek.allocated

MathOptInterfaceMosek.allocatedlist

MathOptInterfaceMosek.allocatevarconstraints

MathOptInterfaceMosek.allocatevariable

MathOptInterfaceMosek.aux_setvardom

MathOptInterfaceMosek.blocksize

MathOptInterfaceMosek.boundflag_all

MathOptInterfaceMosek.boundflag_cone

MathOptInterfaceMosek.boundflag_int

MathOptInterfaceMosek.boundflag_lower

MathOptInterfaceMosek.boundflag_upper

MathOptInterfaceMosek.candelete

MathOptInterfaceMosek.checkconsistency — Function.

Check consistency of the internal structures.

MathOptInterfaceMosek.chgbound

MathOptInterfaceMosek.coneparfromset

MathOptInterfaceMosek.deleteblock — Function.

Move a block to the free list.

MathOptInterfaceMosek.domain_type_mask

MathOptInterfaceMosek.ensurefree — Function.

ensurefree(s::LinkedInts, N :: Int)

Ensure that there are at least N elements free, and allocate as necessary.

MathOptInterfaceMosek.eval

MathOptInterfaceMosek.getconboundlist

MathOptInterfaceMosek.getfreeindexes — Function.

Get a list if the currently free elements.

MathOptInterfaceMosek.getindex — Function.

getindex(s::LinkedInts, id::Int)

Shortcut for getindexes(s, id)[1] when s.size[id] is 1.

MathOptInterfaceMosek.getindexes — Function.

getindexes(s::LinkedInts, id :: Int)

Return the vector of indices for the block id.

MathOptInterfaceMosek.getoneindex

MathOptInterfaceMosek.getsolcode

MathOptInterfaceMosek.getusedindexes — Function.

Get a list if the currently used elements.

MathOptInterfaceMosek.getvarboundlist

MathOptInterfaceMosek.id2vref

MathOptInterfaceMosek.include

MathOptInterfaceMosek.is_conic_set

MathOptInterfaceMosek.is_positivesemidefinite_set

MathOptInterfaceMosek.isvalid

MathOptInterfaceMosek.mosek_block_type_integer

MathOptInterfaceMosek.mosek_block_type_nonneg

MathOptInterfaceMosek.mosek_block_type_nonpos

MathOptInterfaceMosek.mosek_block_type_psd

MathOptInterfaceMosek.mosek_block_type_qcone

MathOptInterfaceMosek.mosek_block_type_range

MathOptInterfaceMosek.mosek_block_type_rqcone

MathOptInterfaceMosek.mosek_block_type_unallocated

MathOptInterfaceMosek.mosek_block_type_zero

MathOptInterfaceMosek.newblock — Function.

newblock(s::LinkedInts, N :: Int)

Add a new block in list idx

MathOptInterfaceMosek.numblocks

MathOptInterfaceMosek.parametrized_task

MathOptInterfaceMosek.parse_parameters

MathOptInterfaceMosek.problemtype_conic

MathOptInterfaceMosek.problemtype_linear

MathOptInterfaceMosek.problemtype_quadratic

MathOptInterfaceMosek.ref2id

MathOptInterfaceMosek.select

MathOptInterfaceMosek.set_internal_name

MathOptInterfaceMosek.solsize

MathOptInterfaceMosek.sympackedLtoU

MathOptInterfaceMosek.sympackedUtoL

MathOptInterfaceMosek.sympackedUtoLidx

MathOptInterfaceMosek.sympackeddim

MathOptInterfaceMosek.sympackedlen

MathOptInterfaceMosek.trimapL

MathOptInterfaceMosek.trimapU