Docstrings

CSTParser.@addctx

CSTParser.@closebrace

CSTParser.@closeparen

CSTParser.@closer — Macro.

@closer ps rule body

Continues parsing closing on rule.

CSTParser.@closesquare

CSTParser.@default — Macro.

@default ps body

Parses the next expression using default closure rules.

CSTParser.@enterctx

CSTParser.@exitctx

CSTParser.@nocloser — Macro.

@nocloser ps rule body

Continues parsing not closing on rule.

CSTParser.@precedence — Macro.

@precedence ps prec body

Continues parsing binary operators until it hits a more loosely binding operator (with precdence lower than prec).

CSTParser.Abstract

CSTParser.AbstractEXPR

CSTParser.AnonFuncOp

CSTParser.ArrowOp

CSTParser.AssignmentOp

CSTParser.BareModule

CSTParser.Begin

CSTParser.BinaryOpCall

CSTParser.BinarySyntaxOpCall

CSTParser.BitShiftOp

CSTParser.Block

CSTParser.Braces

CSTParser.BracesCat

CSTParser.CSTParser

CSTParser.Call

CSTParser.ChainOpCall

CSTParser.Closer

CSTParser.Closer_TMP

CSTParser.ColonOp

CSTParser.ColonOpCall

CSTParser.Comparison

CSTParser.ComparisonOp

CSTParser.Comprehension

CSTParser.ConditionalOp

CSTParser.ConditionalOpCall

CSTParser.Const

CSTParser.Curly

CSTParser.DddotOp

CSTParser.DeclarationOp

CSTParser.DictComprehension

CSTParser.Do

CSTParser.DotOp

CSTParser.EXPR

CSTParser.EmptyWS

CSTParser.EmptyWSToken

CSTParser.Error

CSTParser.ErrorToken

CSTParser.Export

CSTParser.Expr_char

CSTParser.Expr_cmd

CSTParser.Expr_float

CSTParser.Expr_int

CSTParser.Expr_tcmd

CSTParser.FALSE

CSTParser.File

CSTParser.FileH

CSTParser.Filter

CSTParser.Flatten

CSTParser.For

CSTParser.FunctionDef

CSTParser.Generator

CSTParser.Global

CSTParser.GlobalRefDOC

CSTParser.GlobalRefDoc

CSTParser.Hcat

CSTParser.Head

CSTParser.IDENTIFIER

CSTParser.INSTANCE

CSTParser.If

CSTParser.Import

CSTParser.ImportAll

CSTParser.InvisBrackets

CSTParser.KEYWORD

CSTParser.Kw

CSTParser.LITERAL

CSTParser.LazyAndOp

CSTParser.LazyOrOp

CSTParser.LeafNode

CSTParser.Let

CSTParser.Local

CSTParser.LtoR

CSTParser.Macro

CSTParser.MacroCall

CSTParser.MacroName

CSTParser.ModuleH

CSTParser.Mutable

CSTParser.NOTHING

CSTParser.NewLineWS

CSTParser.OPERATOR

CSTParser.Outer

CSTParser.PUNCTUATION

CSTParser.Parameters

CSTParser.ParseState

CSTParser.PipeOp

CSTParser.PlusOp

CSTParser.PowerOp

CSTParser.PrimeOp

CSTParser.Primitive

CSTParser.Project

CSTParser.Quote

CSTParser.Quotenode

CSTParser.RationalOp

CSTParser.Ref

CSTParser.Return

CSTParser.Row

CSTParser.SemiColonWS

CSTParser.StringH

CSTParser.Struct

CSTParser.TRUE

CSTParser.TYPEMAX_INT128_STR

CSTParser.TYPEMAX_INT64_STR

CSTParser.TimesOp

CSTParser.TopLevel

CSTParser.Try

CSTParser.TupleH

CSTParser.TypedComprehension

CSTParser.TypedHcat

CSTParser.TypedVcat

CSTParser.UNICODE_OPS_REVERSE

CSTParser.UnaryOpCall

CSTParser.UnarySyntaxOpCall

CSTParser.Using

CSTParser.Vcat

CSTParser.Vect

CSTParser.WS

CSTParser.WhereOp

CSTParser.WhereOpCall

CSTParser.While

CSTParser._arg_id

CSTParser._get_import_block

CSTParser._parse_ref

CSTParser._unescape_string

CSTParser.accept_comma

CSTParser.accept_end

CSTParser.accept_rbrace

CSTParser.accept_rparen

CSTParser.accept_rsquare

CSTParser.adjustspan

CSTParser.check_base

CSTParser.check_file

CSTParser.check_reformat — Function.

check_reformat()

Reads and parses all files in current directory, applys formatting fixes and checks that the output AST remains the same.

CSTParser.check_span — Function.

check_span(x, neq = [])

Recursively checks whether the span of an expression equals the sum of the span of its components. Returns a vector of failing expressions.

CSTParser.clear_at!

CSTParser.closer

CSTParser.compare — Function.

compare(x,y)

Recursively checks whether two Base.Expr are the same. Returns unequal sub- expressions.

CSTParser.contributes_scope — Function.

contributes_scope(x)

Checks whether the body of x is included in the toplevel namespace.

CSTParser.convert_iter_assign

CSTParser.create_tmp

CSTParser.cst_parsefile

CSTParser.debug

CSTParser.defines_abstract

CSTParser.defines_anon_function

CSTParser.defines_datatype

CSTParser.defines_function

CSTParser.defines_macro

CSTParser.defines_module

CSTParser.defines_mutable

CSTParser.defines_primitive

CSTParser.defines_struct

CSTParser.dropleadlingnewline

CSTParser.eval

CSTParser.expr_import

CSTParser.fix_range

CSTParser.flatten_tuple

CSTParser.flisp_parse

CSTParser.flisp_parsefile

CSTParser.get_arg_name

CSTParser.get_arg_type

CSTParser.get_args

CSTParser.get_body

CSTParser.get_curly_params

CSTParser.get_id — Function.

get_id(x)

Get the IDENTIFIER name of a variable, possibly in the presence of type declaration operators.

CSTParser.get_inner_gen

CSTParser.get_iter

CSTParser.get_name

CSTParser.get_sig — Function.

get_sig(x)

Returns the full signature of function, macro and datatype definitions. Should only be called when has_sig(x) == true.

CSTParser.get_sig_params

CSTParser.get_t — Function.

get_t(x)

Basic inference in the presence of type declarations.

CSTParser.get_where_params

CSTParser.has_sig

CSTParser.include

CSTParser.infer_t

CSTParser.is_and

CSTParser.is_anon_func

CSTParser.is_assignment

CSTParser.is_colon

CSTParser.is_comma

CSTParser.is_cond

CSTParser.is_dddot

CSTParser.is_ddot

CSTParser.is_decl

CSTParser.is_dot

CSTParser.is_elof

CSTParser.is_eq

CSTParser.is_exor

CSTParser.is_float

CSTParser.is_func_call

CSTParser.is_if

CSTParser.is_import

CSTParser.is_importall

CSTParser.is_in

CSTParser.is_integer

CSTParser.is_issubt

CSTParser.is_issupt

CSTParser.is_lbrace

CSTParser.is_lit_string

CSTParser.is_lparen

CSTParser.is_lsquare

CSTParser.is_minus

CSTParser.is_module

CSTParser.is_not

CSTParser.is_nothing

CSTParser.is_number

CSTParser.is_pairarrow

CSTParser.is_plus

CSTParser.is_prime

CSTParser.is_punc

CSTParser.is_range

CSTParser.is_rbrace

CSTParser.is_rparen

CSTParser.is_rsquare

CSTParser.is_star

CSTParser.is_valid_iterator

CSTParser.is_where

CSTParser.isajuxtaposition

CSTParser.isassignment

CSTParser.isbinaryop

CSTParser.isbool

CSTParser.iscomma

CSTParser.isemptyws

CSTParser.isidentifier

CSTParser.isinstance

CSTParser.iskw

CSTParser.ispunctuation

CSTParser.isstring

CSTParser.issyntaxcall

CSTParser.issyntaxunarycall

CSTParser.isunaryandbinaryop

CSTParser.isunaryop

CSTParser.julia_normalization_map

CSTParser.lex_ws_comment — Function.

lex_ws_comment(l::Lexer, c)

Having hit an initial whitespace/comment/semicolon continues collecting similar Chars until they end. Returns a WS token with an indication of newlines/ semicolons. Indicating a semicolons takes precedence over line breaks as the former is equivalent to the former in most cases.

CSTParser.longest_common_prefix

CSTParser.next

CSTParser.non_dotted_op

CSTParser.norm_ast

CSTParser.normalize_julia_identifier

CSTParser.parse — Function.

parse(str, cont = false)

Parses the passed string. If cont is true then will continue parsing until the end of the string returning the resulting expressions in a TOPLEVEL block.

CSTParser.parse_abstract

CSTParser.parse_array — Function.

parse_array(ps)

Having hit '[' return either:

A vect

A vcat

A comprehension

An array (vcat of hcats)

CSTParser.parse_begin

CSTParser.parse_block

CSTParser.parse_braces

CSTParser.parse_call — Function.

parse_call(ps, ret)

Parses a function call. Expects to start before the opening parentheses and is passed the expression declaring the function name, ret.

CSTParser.parse_comma_sep

CSTParser.parse_comp_operator

CSTParser.parse_compound

CSTParser.parse_const

CSTParser.parse_curly — Function.

parse_curly(ps, ret)

Parses the juxtaposition of ret with an opening brace. Parses a comma seperated list.

CSTParser.parse_directory

CSTParser.parse_do

CSTParser.parse_doc

CSTParser.parse_dot_mod

CSTParser.parse_end

CSTParser.parse_export

CSTParser.parse_expression — Function.

parse_expression(ps)

Parses an expression until closer(ps) == true. Expects to enter the ParseState the token before the the beginning of the expression and ends on the last token.

Acceptable starting tokens are:

A keyword

An opening parentheses or brace.

An operator.

An instance (e.g. identifier, number, etc.)

An @.

CSTParser.parse_file

CSTParser.parse_for

CSTParser.parse_function

CSTParser.parse_generator — Function.

parse_generator(ps)

Having hit for not at the beginning of an expression return a generator. Comprehensions are parsed as SQUAREs containing a generator.

CSTParser.parse_global

CSTParser.parse_if — Function.

parse_if(ps, ret, nested=false, puncs=[])

Parse an if block.

CSTParser.parse_imports

CSTParser.parse_iter

CSTParser.parse_kw

CSTParser.parse_let

CSTParser.parse_local

CSTParser.parse_macro

CSTParser.parse_macrocall — Function.

parse_macrocall(ps)

Parses a macro call. Expects to start on the @.

CSTParser.parse_module

CSTParser.parse_mutable

CSTParser.parse_operator

CSTParser.parse_operator_anon_func

CSTParser.parse_operator_colon

CSTParser.parse_operator_cond

CSTParser.parse_operator_dot

CSTParser.parse_operator_eq

CSTParser.parse_operator_power

CSTParser.parse_operator_where

CSTParser.parse_parameters

CSTParser.parse_paren — Function.

parse_paren(ps, ret)

Parses an expression starting with a (.

CSTParser.parse_primitive

CSTParser.parse_quote

CSTParser.parse_ranges

CSTParser.parse_ref — Function.

parse_ref(ps, ret)

Handles cases where an expression - ret - is followed by [. Parses the following bracketed expression and modifies it's .head appropriately.

CSTParser.parse_return

CSTParser.parse_string_or_cmd — Function.

parsestringor_cmd(ps)

When trying to make an INSTANCE from a string token we must check for interpolating opoerators.

CSTParser.parse_struct

CSTParser.parse_try

CSTParser.parse_tuple — Function.

parse_tuple(ps, ret)

ret is followed by a comma so tries to parse the rest of the tuple.

CSTParser.parse_unary — Function.

parse_unary(ps)

Having hit a unary operator at the start of an expression return a call.

CSTParser.parse_unary_colon

CSTParser.parse_while

CSTParser.precedence

CSTParser.read_comment

CSTParser.read_ws

CSTParser.recover_endmarker

CSTParser.rem_call

CSTParser.rem_curly

CSTParser.rem_dddot

CSTParser.rem_decl

CSTParser.rem_invis

CSTParser.rem_kw

CSTParser.rem_subtype

CSTParser.rem_where

CSTParser.remlineinfo! — Function.

remlineinfo!(x)

Removes line info expressions. (i.e. Expr(:line, 1))

CSTParser.sized_uint_literal

CSTParser.sized_uint_oct_literal

CSTParser.skip_to_nl

CSTParser.span

CSTParser.speed_test

CSTParser.str_value

CSTParser.term_c

CSTParser.test_find

CSTParser.test_order

CSTParser.tostr

CSTParser.trailing_ws_length

CSTParser.update_from_tmp!

CSTParser.update_span!

CSTParser.update_to_default!

CSTParser.utf8proc_map_custom

CSTParser.val

CSTParser.wstype

CSTParser.x_Cmd

CSTParser.x_Str