Docstrings

Turing.+

Turing.-

Turing.@VarName — Macro.

Usage: @VarName x[1,2][1+5][45][3] return: (:x,[1,2],6,45,3)

Turing.@model — Macro.

@model(name, fbody)

Macro to specify a probabilistic model.

Example:

@model Gaussian(x) = begin
    s ~ InverseGamma(2,3)
    m ~ Normal(0,sqrt.(s))
    for i in 1:length(x)
        x[i] ~ Normal(m, sqrt.(s))
    end
    return (s, m)
end

Compiler design: sample(fname(x,y), sampler).

fname(x=nothing,y=nothing; compiler=compiler) = begin
    ex = quote
        # Pour in kwargs for those args where value != nothing.
        fname_model(vi::VarInfo, sampler::Sampler; x = x, y = y) = begin
            vi.logp = zero(Real)
          
            # Pour in model definition.
            x ~ Normal(0,1)
            y ~ Normal(x, 1)
            return x, y
        end
    end
    return Main.eval(ex)
end

Turing.@sym_str

Turing.@~ — Macro.

macro: @~ var Distribution()

Tilde notation macro. This macro constructs Turing.observe or Turing.assume calls depending on the left-hand argument. Note that the macro is interconnected with the @model macro and assumes that a compiler struct is available.

Example:

@~ x Normal()

Turing.ADBACKEND

Turing.ADSAFE

Turing.AbstractMixtureModel

Turing.AbstractMvNormal

Turing.AbstractSampler

Turing.AnySampler

Turing.Arcsine

Turing.Bernoulli

Turing.Beta

Turing.BetaBinomial

Turing.BetaPrime

Turing.Binomial

Turing.Biweight

Turing.CACHEIDCS

Turing.CACHERANGES

Turing.CACHERESET

Turing.CHUNKSIZE

Turing.CSMC

Turing.Categorical

Turing.Cauchy

Turing.Chain — Type.

Chain(weight::Float64, value::Array{Sample})

A wrapper of output trajactory of samplers.

Example:

# Define a model
@model xxx begin
  ...
  return(mu,sigma)
end

# Run the inference engine
chain = sample(xxx, SMC(1000))

chain[:logevidence]   # show the log model evidence
chain[:mu]            # show the weighted trajactory for :mu
chain[:sigma]         # show the weighted trajactory for :sigma
mean(chain[:mu])      # find the mean of :mu
mean(chain[:sigma])   # find the mean of :sigma

Turing.Chains

Turing.Chi

Turing.Chisq

Turing.Continuous

Turing.ContinuousDistribution

Turing.ContinuousMatrixDistribution

Turing.ContinuousMultivariateDistribution

Turing.ContinuousUnivariateDistribution

Turing.Cosine

Turing.DEFAULT_ADAPT_CONF_TYPE

Turing.DiagNormal

Turing.DiagNormalCanon

Turing.Dirichlet

Turing.DirichletMultinomial

Turing.Discrete

Turing.DiscreteDistribution

Turing.DiscreteMatrixDistribution

Turing.DiscreteMultivariateDistribution

Turing.DiscreteUniform

Turing.DiscreteUnivariateDistribution

Turing.Distribution

Turing.Distributions

Turing.DoubleExponential

Turing.DynamicNUTS

Turing.EdgeworthMean

Turing.EdgeworthSum

Turing.EdgeworthZ

Turing.EmpiricalUnivariateDistribution

Turing.Epanechnikov

Turing.Erlang

Turing.Estimator

Turing.Exponential

Turing.FDist

Turing.FisherNoncentralHypergeometric

Turing.Flat

Turing.FlatPos

Turing.Frechet

Turing.FullNormal

Turing.FullNormalCanon

Turing.Gamma

Turing.GeneralizedExtremeValue

Turing.GeneralizedPareto

Turing.Geometric

Turing.Gibbs — Type.

Gibbs(n_iters, alg_1, alg_2)

Compositional MCMC interface.

Example:

alg = Gibbs(1000, HMC(1, 0.2, 3, :v1), PG(20, 1, :v2))

Turing.GibbsComponent

Turing.Gumbel

Turing.HMC — Type.

HMC(n_iters::Int, epsilon::Float64, tau::Int)

Hamiltonian Monte Carlo sampler.

Usage:

HMC(1000, 0.05, 10)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
    s ~ InverseGamma(2,3)
    m ~ Normal(0, sqrt(s))
    x[1] ~ Normal(m, sqrt(s))
    x[2] ~ Normal(m, sqrt(s))
    return s, m
end

sample(gdemo([1.5, 2]), HMC(1000, 0.05, 10))

Turing.HMCDA — Type.

HMCDA(n_iters::Int, n_adapt::Int, delta::Float64, lambda::Float64)

Hamiltonian Monte Carlo sampler wiht Dual Averaging algorithm.

Usage:

HMCDA(1000, 200, 0.65, 0.3)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0, sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), HMCDA(1000, 200, 0.65, 0.3))

Turing.Hamiltonian

Turing.HamiltonianRobustInit — Type.

Robust initialization method for model parameters in Hamiltonian samplers.

Turing.Hypergeometric

Turing.IArray

Turing.IPMCMC — Type.

IPMCMC(n_particles::Int, n_iters::Int, n_nodes::Int, n_csmc_nodes::Int)

Particle Gibbs sampler.

Usage:

IPMCMC(100, 100, 4, 2)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), IPMCMC(100, 100, 4, 2))

Turing.IS — Type.

IS(n_particles::Int)

Importance sampling algorithm object.

n_particles is the number of particles to use

Usage:

IS(1000)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
    s ~ InverseGamma(2,3)
    m ~ Normal(0,sqrt.(s))
    x[1] ~ Normal(m, sqrt.(s))
    x[2] ~ Normal(m, sqrt.(s))
    return s, m
end

sample(gdemo([1.5, 2]), IS(1000))

Turing.InferenceAlgorithm

Turing.InverseGamma

Turing.InverseGaussian

Turing.InverseWishart

Turing.IsoNormal

Turing.IsoNormalCanon

Turing.KSDist

Turing.KSOneSided

Turing.Kolmogorov

Turing.Laplace

Turing.Levy

Turing.LocationScale

Turing.LogNormal

Turing.Logistic

Turing.MCMCChain

Turing.MH — Type.

MH(n_iters::Int)

Metropolis-Hasting sampler.

Usage:

MH(100, (:m, (x) -> Normal(x, 0.1)))

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), MH(1000, (:m, (x) -> Normal(x, 0.1)), :s)))

Turing.MLEstimator

Turing.MatrixDistribution

Turing.Matrixvariate

Turing.MixtureModel

Turing.Multinomial

Turing.Multivariate

Turing.MultivariateDistribution

Turing.MultivariateMixture

Turing.MultivariateNormal

Turing.MvLogNormal

Turing.MvNormal

Turing.MvNormalCanon

Turing.MvNormalKnownCov

Turing.MvTDist

Turing.NUTS — Type.

NUTS(n_iters::Int, n_adapt::Int, delta::Float64)

No-U-Turn Sampler (NUTS) sampler.

Usage:

NUTS(1000, 200, 0.6j_max)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0, sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.j_max, 2]), NUTS(1000, 200, 0.6j_max))

Turing.NegativeBinomial

Turing.NonMatrixDistribution

Turing.NoncentralBeta

Turing.NoncentralChisq

Turing.NoncentralF

Turing.NoncentralHypergeometric

Turing.NoncentralT

Turing.Normal

Turing.NormalCanon

Turing.NormalInverseGaussian

Turing.PG — Type.

PG(n_particles::Int, n_iters::Int)

Particle Gibbs sampler.

Usage:

PG(100, 100)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0, sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), PG(100, 100))

Turing.PIMH

Turing.PMMH — Type.

PMMH(n_iters::Int, smc_alg:::SMC, parameters_algs::Tuple{MH})

Particle independant Metropolis–Hastings and Particle marginal Metropolis–Hastings samplers.

Usage:

alg = PMMH(100, SMC(20, :v1), MH(1,:v2))
alg = PMMH(100, SMC(20, :v1), MH(1,(:v2, (x) -> Normal(x, 1))))

Turing.PROGRESS

Turing.Pareto

Turing.Particle

Turing.ParticleContainer — Type.

Data structure for particle filters

effectiveSampleSize(pc :: ParticleContainer)

normalise!(pc::ParticleContainer)

consume(pc::ParticleContainer): return incremental likelihood

Turing.Poisson

Turing.PoissonBinomial

Turing.QQPair

Turing.Rayleigh

Turing.RealInterval

Turing.SGHMC — Type.

SGHMC(n_iters::Int, learning_rate::Float64, momentum_decay::Float64)

Stochastic Gradient Hamiltonian Monte Carlo sampler.

Usage:

SGHMC(1000, 0.01, 0.1)

Example:

@model example begin
  ...
end

sample(example, SGHMC(1000, 0.01, 0.1))

Turing.SGLD — Type.

SGLD(n_iters::Int, step_size::Float64)

Stochastic Gradient Langevin Dynamics sampler.

Usage:

SGLD(1000, 0.5)

Example:

@model example begin
  ...
end

sample(example, SGLD(1000, 0.5))

Turing.SMC — Type.

SMC(n_particles::Int)

Sequential Monte Carlo sampler.

Usage:

SMC(1000)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0, sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), SMC(1000))

Turing.STAN_DEFAULT_ADAPT_CONF

Turing.Sample

Turing.SampleFromPrior

Turing.Sampleable

Turing.Sampler — Type.

Sampler{T}

Generic interface for implementing inference algorithms. An implementation of an algorithm should include the following:

A type specifying the algorithm and its parameters, derived from InferenceAlgorithm

A method of sample function that produces results of inference, which is where actual inference happens.

Turing translates models to chunks that call the modelling functions at specified points. The dispatch is based on the value of a sampler variable. To include a new inference algorithm implements the requirements mentioned above in a separate file, then include that file at the end of this one.

Turing.Semicircle

Turing.Skellam

Turing.SufficientStats

Turing.SymTriangularDist

Turing.TArray

Turing.TDist

Turing.Trace

Turing.Transformable

Turing.TriangularDist

Turing.Triweight

Turing.Truncated

Turing.TruncatedNormal

Turing.Turing

Turing.Uniform

Turing.Univariate

Turing.UnivariateDistribution

Turing.UnivariateGMM

Turing.UnivariateMixture

Turing.ValueSupport

Turing.VarEstimator

Turing.VarReplay

Turing.VariateForm

Turing.VonMises

Turing.VonMisesFisher

Turing.WalleniusNoncentralHypergeometric

Turing.WarmUpManager

Turing.Weibull

Turing.Wishart

Turing.ZeroMeanDiagNormal

Turing.ZeroMeanDiagNormalCanon

Turing.ZeroMeanFullNormal

Turing.ZeroMeanFullNormalCanon

Turing.ZeroMeanIsoNormal

Turing.ZeroMeanIsoNormalCanon

Turing.__init__

Turing.__inits__

Turing._build_tree — Function.

function buildtree(θ::T, r::AbstractVector, logu::AbstractFloat, v::Int, j::Int, ϵ::AbstractFloat, H0::AbstractFloat,ljfunc::Function, gradfunc::Function, stds::AbstractVector; Δ_max::AbstractFloat=1000) where {T<:Union{Vector,SubArray}}

Recursively build balanced tree.

Ref: Algorithm 6 on http://www.stat.columbia.edu/~gelman/research/published/nuts.pdf

Arguments:

θ : model parameter

r : momentum variable

logu : slice variable (in log scale)

v : direction ∈ {-1, 1}

j : depth of tree

ϵ : leapfrog step size

H0 : initial H

lj_func : function for log-joint

grad_func : function for the gradient of log-joint

stds : pre-conditioning matrix

Δ_max : threshold for exploeration error tolerance

Turing._find_H

Turing._hmc_step

Turing._leapfrog

Turing._nuts_step — Function.

function nutsstep(θ::T, ϵ::AbstractFloat, ljfunc::Function, gradfunc::Function, stds::AbstractVector; jmax::Int=jmax) where {T<:Union{AbstractVector,SubArray}}

Perform one NUTS step.

Ref: Algorithm 6 on http://www.stat.columbia.edu/~gelman/research/published/nuts.pdf

Arguments:

θ : model parameter

ϵ : leapfrog step size

lj_func : function for log-joint

grad_func : function for the gradient of log-joint

stds : pre-conditioning matrix

j_max : maximum expanding of doubling tree

Turing._sample_momentum

Turing.adapt!

Turing.adapt_step_size!

Turing.add_sample!

Turing.assume

Turing.autcorplot

Turing.auto_tune_chunk_size!

Turing.autocorplot

Turing.autocorplot!

Turing.binaryentropy

Turing.canonform

Turing.ccdf

Turing.cdf

Turing.cf

Turing.cgf

Turing.circvar

Turing.component

Turing.components

Turing.componentwise_logpdf

Turing.componentwise_pdf

Turing.compute_next_window

Turing.concentration

Turing.consume

Turing.cor

Turing.corner

Turing.corner!

Turing.cov

Turing.cquantile

Turing.cumulant

Turing.current_trace

Turing.data — Function.

data(dict::Dict, keys::Vector{Symbol})

Construct a tuple with values filled according to dict and keys according to keys.

Turing.data_insertion

Turing.densityplot

Turing.describe

Turing.dim

Turing.discretediag

Turing.dof

Turing.effectiveSampleSize

Turing.entropy

Turing.estimate

Turing.eval

Turing.expected_logdet

Turing.failprob

Turing.find_H

Turing.find_good_eps

Turing.fit

Turing.fit_map

Turing.fit_map!

Turing.fit_mle

Turing.fit_mle!

Turing.flatten

Turing.flatten!

Turing.fork

Turing.forkr

Turing.freecumulant

Turing.gelmandiag

Turing.gen_grad_func — Function.

gen_grad_func(vi::VarInfo, sampler::Sampler, model)

Generate a function that takes a vector of reals θ and compute the logpdf and gradient at θ for the model specified by (vi, sampler, model).

Turing.gen_lj_func — Function.

gen_lj_func(vi::VarInfo, sampler::Sampler, model)

Generate a function that takes θ and returns logpdf at θ for the model specified by (vi, sampler, model).

Turing.gen_log_func — Function.

gen_log_func(sampler::Sampler)

Generate a function that takes no argument and performs logging for the number of leapfrog steps used in sampler.

Turing.gen_rev_func — Function.

genrevfunc(vi::VarInfo, sampler::Sampler)

Generate a function on (θ, logp) that sets the variables referenced by sampler to θ and the current vi.logp to logp.

Turing.generate_assume — Function.

generate_assume(variable, distribution, syms)

Generate an assume expression.

Turing.generate_observe — Function.

generate_observe(observation, distribution)

Generate an observe expression.

Turing.get_var

Turing.getindex

Turing.getjuliatype

Turing.getsample

Turing.getvsym

Turing.gewekediag

Turing.gradient_forward — Function.

gradientforward( θ::AbstractVector{<:Real}, vi::VarInfo, model::Function, spl::Union{Nothing, Sampler}=nothing, chunksize::Int=CHUNKSIZE[], )

Computes the gradient of the log joint of θ for the model specified by (vi, spl, model) using forwards-mode AD from ForwardDiff.jl.

Turing.gradient_reverse — Function.

gradient_reverse( θ::AbstractVector{<:Real}, vi::VarInfo, model::Function, sampler::Union{Nothing, Sampler}=nothing, )

Computes the gradient of the log joint of θ for the model specified by (vi, sampler, model) using reverse-mode AD from Flux.jl.

Turing.gradlogpdf

Turing.hasfinitesupport

Turing.heideldiag

Turing.histogramplot

Turing.in_adaptation

Turing.include

Turing.increase_logevidence

Turing.increase_logweight

Turing.ind2sub

Turing.init

Turing.init_warm_up_params

Turing.inittrans

Turing.insupport

Turing.invcov

Turing.invlogccdf

Turing.invlogcdf

Turing.invscale

Turing.is_window_end

Turing.isbounded

Turing.isleptokurtic

Turing.islowerbounded

Turing.ismesokurtic

Turing.isplatykurtic

Turing.isprobvec

Turing.isupperbounded

Turing.kde

Turing.kurtosis

Turing.leapfrog

Turing.localcopy

Turing.location

Turing.location!

Turing.logccdf

Turing.logcdf

Turing.logdetcov

Turing.loglikelihood

Turing.logpdf

Turing.logpdf!

Turing.mean

Turing.meandir

Turing.meanform

Turing.meanlogx

Turing.meanplot

Turing.meanplot!

Turing.median

Turing.mgf

Turing.mh_accept — Function.

mh_accept(H::Real, H_new::Real)
mh_accept(H::Real, H_new::Real, log_proposal_ratio::Real)

Peform MH accept criteria. Returns a boolean for whether or not accept and the acceptance ratio in log space.

Turing.mixeddensity

Turing.mixeddensity!

Turing.mixeddensityplot

Turing.mode

Turing.modes

Turing.moment

Turing.ncategories

Turing.ncomponents

Turing.nsamples

Turing.ntrials

Turing.observe

Turing.params

Turing.params!

Turing.partype

Turing.pdf

Turing.plot

Turing.probs

Turing.probval

Turing.produce

Turing.propose

Turing.qqbuild

Turing.quantile

Turing.rafterydiag

Turing.randcat

Turing.randrealuni

Turing.rate

Turing.reconstruct

Turing.reconstruct!

Turing.require_gradient

Turing.require_particles

Turing.resample

Turing.resample!

Turing.resampleMultinomial

Turing.resampleResidual

Turing.resampleStratified

Turing.resampleSystematic

Turing.reset!

Turing.restart_da

Turing.resume

Turing.runmodel!

Turing.sample

Turing.sample!

Turing.sample_momentum

Turing.sampler

Turing.save!

Turing.scale

Turing.scale!

Turing.setadbackend

Turing.setadsafe

Turing.setchunksize

Turing.setindex!

Turing.setkwargs

Turing.shape

Turing.skewness

Turing.span

Turing.sqmahal

Turing.sqmahal!

Turing.std

Turing.stdlogx

Turing.step

Turing.succprob

Turing.suffstats

Turing.support

Turing.test_distr

Turing.test_samples

Turing.tilde

Turing.traceplot

Turing.traceplot!

Turing.translate

Turing.translate!

Turing.turnprogress

Turing.tzeros

Turing.update_da_μ

Turing.update_pre_cond!

Turing.var

Turing.var_tuple

Turing.varlogx

Turing.vectorize

Turing.verifygrad

Turing.weights

Turing.wsample

Turing.wsample!