Docstrings

MLLabelUtils.BinaryLabelEncoding

MLLabelUtils.LabelEnc — Module.

The submodule MLLabelUtils.LabelEnc contains a selection of popular label encodings:

Strictly binary label encodings:

LabelEnc.FuzzyBinary

LabelEnc.TrueFalse

LabelEnc.ZeroOne

LabelEnc.MarginBased

Multiclass label encodings:

LabelEnc.Indices

LabelEnc.NativeLabels

LabelEnc.OneOfK

Multiclass to binary:

LabelEnc.OneVsRest

MLLabelUtils.LabelEncoding

MLLabelUtils.MLLabelUtils

MLLabelUtils.MatrixLabelEncoding

MLLabelUtils.ObsDim

MLLabelUtils.VectorLabelEncoding

MLLabelUtils._ambiguous

MLLabelUtils._arrange_label

MLLabelUtils._array_type

MLLabelUtils._lm

MLLabelUtils._sign

MLLabelUtils.classify — Function.

classify(x, encoding)

Returns the classified version of x given the encoding. Which means that if x can be interpreted as a positive label, the positive label of encoding is returned; the negative otherwise.

julia> classify(0.6, LabelEnc.ZeroOne(UInt8))
0x01

julia> classify(0.4, LabelEnc.ZeroOne(UInt8))
0x00

julia> classify([0.1,0.2,0.6,0.1], LabelEnc.OneOfK)
3

MLLabelUtils.classify! — Function.

classify!(out, x, encoding)

Same as classify, but uses out to store the result.

julia> buffer = zeros(2);
julia> classify!(buffer, [0.4,0.6], LabelEnc.ZeroOne)
2-element Array{Float64,1}:
 0.0
 1.0

MLLabelUtils.convertlabel — Function.

convertlabel(new_encoding, x, [old_encoding])

Converts the given value/array x from the old_encoding into the new_encoding. Note that if old_encoding is not specified it will be derived automaticaly using labelenc.

julia> convertlabel(LabelEnc.MarginBased, [0, 1, 1, 0, 0])
5-element Array{Int64,1}:
 -1
  1
  1
 -1
 -1

julia> convertlabel([:yes,:no], [0, 1, 1, 0, 0])
5-element Array{Symbol,1}:
 :no
 :yes
 :yes
 :no
 :no

For more information on the available encodings, see ?LabelEnc.

convertlabel(new_encoding, x, [old_encoding], [obsdim])

When working with OneOfK one can additionally specifify which dimension of the array denotes the observations using obsdim

julia> convertlabel(LabelEnc.OneOfK, [0, 1, 1, 0, 0], obsdim = 2)
2×5 Array{Int64,2}:
 0  1  1  0  0
 1  0  0  1  1

MLLabelUtils.convertlabel!

MLLabelUtils.convertlabelview — Function.

convertlabel(new_encoding, vec::AbstractVector, [old_encoding]) -> (Readonly)MappedArray

Creates a lazy view into vec that makes it look like it is in the encoding specified by new_encoding, while it is actually preserved as being of old_encoding.

This method only works for label-encodings that are vector-based (i.e. pretty much all but OneOfK). The resulting MappedArray will be writeable unless old_encoding is of type OneVsRest, in which case the result will be a ReadonlyMappedArray.

MLLabelUtils.eval

MLLabelUtils.include

MLLabelUtils.ind2label — Function.

ind2label(index, encoding)

Converts the given index into the corresponding label defined by the encoding. Note that in the binary case, index == 1 represents the positive label and index == 2 the negative label.

julia> ind2label(1, LabelEnc.ZeroOne(Float32))
1.0f0

julia> ind2label(2, LabelEnc.ZeroOne(Float32))
0.0f0

MLLabelUtils.islabelenc — Function.

islabelenc(obj, encoding) -> Bool

Checks is the given object obj can be described as being produced by the given encoding in which case the function returns true, or false otherwise.

julia> islabelenc([1,0,1], LabelEnc.ZeroOne)
true

julia> islabelenc([1,-1,1], LabelEnc.ZeroOne)
false

MLLabelUtils.isneglabel — Function.

isneglabel(x, encoding) -> Bool

Checks if the given value x can be interpreted as the negative label given the encoding. This function takes potential classification rules into account.

julia> isneglabel(0.6, LabelEnc.ZeroOne(0.5))
false

julia> isneglabel(0.4, LabelEnc.ZeroOne(0.5))
true

julia> isneglabel(:b, LabelEnc.NativeLabels([:a,:b]))
true

MLLabelUtils.isposlabel — Function.

isposlabel(x, encoding) -> Bool

Checks if the given value x can be interpreted as the positive label given the encoding. This function takes potential classification rules into account.

julia> isposlabel(0.6, LabelEnc.ZeroOne(0.5))
true

julia> isposlabel(0.4, LabelEnc.ZeroOne(0.5))
false

julia> isposlabel(:b, LabelEnc.NativeLabels([:a,:b]))
false

MLLabelUtils.label — Function.

label(obj) -> Vector

Returns the labels represented in the given object obj. Note that the order of the labels matters. In the case of two labels, the first element represents the positive label and the second element the negative label.

julia> label([:yes,:no,:yes,:yes])
2-element Array{Symbol,1}:
 :yes
 :no

julia> label(LabelEnc.ZeroOne())
2-element Array{Float64,1}:
 1.0
 0.0

MLLabelUtils.label2ind — Function.

label2ind(label, encoding) -> Int

Converts the given label into the corresponding index defined by the encoding. Note that in the binary case, the positive label will result in the index 1 and the negative label in the index 2 respectively

julia> label2ind(:no, LabelEnc.NativeLabels([:yes,:no]))
2

julia> label2ind(1, LabelEnc.MarginBased())
1

julia> label2ind(-1, LabelEnc.MarginBased())
2

MLLabelUtils.labelenc — Function.

labelenc(obj) -> LabelEncoding

Tries to determine the most approriate label-encoding to describe the given object obj based on the result of label(obj). Note that in most cases this function is not typestable.

julia> labelenc([:yes,:no,:no,:yes,:maybe])
MLLabelUtils.LabelEnc.NativeLabels{Symbol,3}(Symbol[:yes,:no,:maybe],Dict(:yes=>1,:maybe=>3,:no=>2))

julia> labelenc([1,0,0,1,0,1])
MLLabelUtils.LabelEnc.ZeroOne{Int64,Float64}(0.5)

julia> labelenc(Int8[1,-1,-1,1,-1,1])
MLLabelUtils.LabelEnc.MarginBased{Int8}()

MLLabelUtils.labelfreq — Function.

labelfreq(obj) -> Dict

Computes the absolute frequencies for each label in obj.

julia> labelfreq([0, 1, 1, 0, 0])
Dict{Int64,Int64} with 2 entries:
  0 => 3
  1 => 2

MLLabelUtils.labelfreq! — Function.

labelfreq!(dict, obj) -> Dict

updates the given label-frequency-map dict with the absolute frequencies for each label in obj

julia> ld = labelfreq([0, 1, 1, 0, 0])
Dict{Int64,Int64} with 2 entries:
  0 => 3
  1 => 2

julia> labelfreq!(ld, [1,0,0])
Dict{Int64,Int64} with 2 entries:
  0 => 5
  1 => 3

MLLabelUtils.labelmap — Function.

labelmap(obj) -> Dict

Computes a mapping from the labels in obj to all the individual element-indices in obj that correspond to that label

julia> labelmap([0, 1, 1, 0, 0])
Dict{Int64,Array{Int64,1}} with 2 entries:
  0 => [1,4,5]
  1 => [2,3]

MLLabelUtils.labelmap! — Function.

labelmap!(dict, idx, elem) -> Dict

Updates the given label-map dict with the new element elem, which is assumed to be associated with the index idx.

julia> lm = labelmap([0, 1, 1, 0, 0])
Dict{Int64,Array{Int64,1}} with 2 entries:
  0 => [1,4,5]
  1 => [2,3]

julia> labelmap!(lm, 6, 0)
Dict{Int64,Array{Int64,1}} with 2 entries:
  0 => [1,4,5,6]
  1 => [2,3]

julia> labelmap!(lm, 7:8, [1,0])
Dict{Int64,Array{Int64,1}} with 2 entries:
  0 => [1,4,5,6,8]
  1 => [2,3,7]

MLLabelUtils.labelmap2vec — Function.

labelmap2vec(dict) -> Vector

Inverse function of labelmap. Computes an array of labels by element-wise traversal of the entries in dict.

julia> labelvec = [:yes,:no,:no,:yes,:yes]

julia> lm = labelmap(labelvec)
Dict{Symbol,Array{Int64,1}} with 2 entries:
    :yes => [1, 4, 5]
    :no  => [2, 3]

julia> labelmap2vec(lm)
5-element Array{Symbol,1}:
    :yes
    :no
    :no
    :yes
    :yes

MLLabelUtils.labeltype — Function.

labeltype(::Type{<:LabelEncoding}) -> DataType

Determine the type of the labels represented by the label encoding T

MLLabelUtils.neglabel — Function.

neglabel(encoding)

If the encoding is binary it will return the negative label of it. The function will throw an error otherwise.

julia> neglabel(LabelEnc.MarginBased(Float32))
-1.0f0

MLLabelUtils.nlabel — Function.

nlabel(obj) -> Int

Returns the number of labels represented in the given object obj.

julia> nlabel([:yes,:no,:yes,:yes])
2

MLLabelUtils.poslabel — Function.

poslabel(encoding)

If the encoding is binary it will return the positive label of it. The function will throw an error otherwise.

julia> poslabel(LabelEnc.MarginBased(Float32))
1.0f0

MLLabelUtils.standardize_label