Docstrings

ValueHistories.@trace — Macro.

Easily add to a MVHistory object tr.

Example:

using ValueHistories, OnlineStats
v = Variance(BoundedEqualWeight(30))
tr = MVHistory()
for i=1:100
    r = rand()
    fit!(v,r)
    μ,σ = mean(v),std(v)

    # add entries for :r, :μ, and :σ using their current values
    @trace tr i r μ σ
end

ValueHistories.History

ValueHistories.MVHistory

ValueHistories.MultivalueHistory

ValueHistories.QHistory

ValueHistories.UnivalueHistory

ValueHistories.ValueHistories

ValueHistories.ValueHistory

ValueHistories._filter_plotable_histories

ValueHistories._is_plotable_history

ValueHistories.eval

ValueHistories.include

ValueHistories.increment! — Function.

increment!(trace, iter, val)

Increments the value for a given iteration if it exists, otherwise adds the iteration with an ordinary push.

increment!(trace, key, iter, val)

Increments the value for a given key and iteration if it exists, otherwise adds the key/iteration pair with an ordinary push.