Docstrings

Intervals...

Intervals.AbstractInterval

Intervals.AnchoredInterval — Type.

AnchoredInterval{P, T}(anchor::T, [inclusivity::Inclusivity]) where {P, T} -> AnchoredInterval{P, T}
AnchoredInterval{P, T}(anchor::T, [closed_left::Bool, closed_right::Bool]) where {P, T} -> AnchoredInterval{P, T}

AnchoredInterval is a subtype of AbstractInterval that represents a non-iterable range or span of values defined not by two endpoints but instead by a single anchor point and the value type P which represents the size of the range. When P is positive, the anchor represents the lesser endpoint (the beginning of the range); when P is negative, the anchor represents the greater endpoint (the end of the range).

The interval represented by an AnchoredInterval value may be closed (both endpoints are included in the interval), open (neither endpoint is included), or half-open. This openness is defined by an Inclusivity value, which defaults to half-open (with the lesser endpoint included for positive values of P and the greater endpoint included for negative values).

Why?

AnchoredIntervals are most useful in cases where a single value is used to stand in for a range of values. This happens most often with dates and times, where "HE15" is often used as shorthand for (14:00..15:00].

To this end, HourEnding is a type alias for AnchoredInterval{Hour(-1)}. Similarly, HourBeginning is a type alias for AnchoredInterval{Hour(1)}.

Rounding

While the user may expect an HourEnding or HourBeginning value to be anchored to a specific hour, the constructor makes no guarantees that the anchor provided is rounded:

julia> HourEnding(DateTime(2016, 8, 11, 2, 30))
HourEnding{DateTime}(2016-08-11T02:30:00, Inclusivity(false, true))

The HE and HB pseudoconstructors round the input up or down to the nearest hour, as appropriate:

julia> HE(DateTime(2016, 8, 11, 2, 30))
HourEnding{DateTime}(2016-08-11T03:00:00, Inclusivity(false, true))

julia> HB(DateTime(2016, 8, 11, 2, 30))
HourBeginning{DateTime}(2016-08-11T02:00:00, Inclusivity(true, false))

Example

julia> AnchoredInterval{Hour(-1)}(DateTime(2016, 8, 11, 12))
HourEnding{DateTime}(2016-08-11T12:00:00, Inclusivity(false, true))

julia> AnchoredInterval{Day(1)}(DateTime(2016, 8, 11))
AnchoredInterval{1 day, DateTime}(2016-08-11T00:00:00, Inclusivity(true, false))

julia> AnchoredInterval{Minute(5)}(DateTime(2016, 8, 11, 12, 30), true, true)
AnchoredInterval{5 minutes, DateTime}(2016-08-11T12:30:00, Inclusivity(true, true))

See also: Interval, Inclusivity, HE, HB

Intervals.Beginning

Intervals.Direction

Intervals.Ending

Intervals.Endpoint

Intervals.HB — Function.

HB(anchor, args...) -> HourBeginning

HB is a pseudoconstructor for HourBeginning that rounds the anchor provided down to the nearest hour.

Intervals.HE — Function.

HE(anchor, args...) -> HourEnding

HE is a pseudoconstructor for HourEnding that rounds the anchor provided up to the nearest hour.

Intervals.HourBeginning — Type.

HourBeginning{T<:TimeType} <: AbstractInterval{T}

A type alias for AnchoredInterval{Hour(1), T} which is used to denote a 1-hour period of time which begins at a time instant (of type T).

Intervals.HourEnding — Type.

HourEnding{T<:TimeType} <: AbstractInterval{T}

A type alias for AnchoredInterval{Hour(-1), T} which is used to denote a 1-hour period of time which ends at a time instant (of type T).

Intervals.Inclusivity — Type.

Inclusivity(first::Bool, last::Bool) -> Inclusivity

Defines whether an AbstractInterval is open, half-open, or closed.

Intervals.Interval — Type.

Interval(first, last, [inclusivity::Inclusivity]) -> Interval
Interval(first, last, [closed_left::Bool, closed_right::Bool]) -> Interval

An Interval represents a non-iterable range or span of values (non-interable because, unlike a StepRange, no step is defined).

An Interval can be closed (both first and last are included in the interval), open (neither first nor last are included), or half-open. This openness is defined by an Inclusivity value, which defaults to closed.

Example

julia> i = Interval(0, 100, true, false)
Interval{Int64}(0, 100, Inclusivity(true, false))

julia> in(0, i)
true

julia> in(50, i)
true

julia> in(100, i)
false

julia> intersect(Interval(0, 25, false, false), Interval(20, 50, true, true)
Interval{Int64}(20, 25, Inclusivity(true, false))

Infix Constructor: ..

A closed Interval can be constructed with the .. infix constructor:

julia> Dates.today() - Dates.Week(1) .. Dates.today()
Interval{Date}(2018-01-24, 2018-01-31, Inclusivity(true, true))

Note on Ordering

The Interval constructor will compare first and last; if it finds that first > last, they will be reversed to ensure that first < last. This simplifies calls to in and intersect:

julia> i = Interval(Date(2016, 8, 11), Date(2013, 2, 13), false, true)
Interval{Date}(2013-02-13, 2016-08-11, Inclusivity(true, false))

Note that the Inclusivity value is also reversed in this case.

See also: AnchoredInterval, Inclusivity

Intervals.Intervals

Intervals.Left

Intervals.LeftEndpoint

Intervals.Right

Intervals.RightEndpoint

Intervals.anchor

Intervals.canonicalize

Intervals.contiguous

Intervals.description

Intervals.eval

Intervals.first

Intervals.greater_than_disjoint

Intervals.include

Intervals.inclusivity

Intervals.isclosed

Intervals.last

Intervals.less_than_disjoint

Intervals.merge

Intervals.overlaps

Intervals.prefix

Intervals.span

Intervals.superset — Function.

superset(intervals::AbstractArray{<:AbstractInterval}) -> Interval

Create the smallest single interval which encompasses all of the provided intervals.

Intervals.time_string

Base.union — Function.

union(intervals::AbstractVector{<:AbstractInterval})

Flattens a vector of overlapping intervals into a new, smaller vector containing only non-overlapping intervals.

Base.union! — Function.

union!(intervals::AbstractVector{<:Union{Interval, AbstractInterval}})

Flattens a vector of overlapping intervals in-place to be a smaller vector containing only non-overlapping intervals.

Intervals.≪

Intervals.≫

Intervals.⊆

Intervals.⊇

Intervals.⊈

Intervals.⊉