Docstrings

NRRD.IntFloat

NRRD.NRRD

NRRD.PTuple

NRRD.QString

NRRD.StringPTuple

NRRD.UnknownColor — Type.

UnknownColor{T,N}

An unknown Color. This type gets returned when one of the "kind" settings is "3-color" or "4-color".

NRRD.VTuple

NRRD._header_eltype!

NRRD.alloctype

NRRD.arraytype — Function.

arraytype(filename)

Parse NRRD header and calls arraytype!(header, version). See arraytype! for information about the return values.

NRRD.arraytype! — Function.

arraytype!(header, version) -> T, axs, perm, need_bswap

Analyze the header dictionary to extract the element-type T, size or axes information axs, the permutation perm (if any) that julia should use for "wrapping" the read data, and a boolean need_bswap indicating whether the data need to be byte-swapped (to account for differences in endianness). T includes any color information (in which case a dimension of the array will be "consumed"). axs will be a Dims-tuple in simple cases, or an Axes tuple (from AxisArrays.jl) if dimensions are labeled or have their spatial information (pixelspacing, spacedirections, etc) specified. perm is the permutation needed to move the color data to the first dimension, or an empty tuple if no permutation is required.

This function may modify the header dictionary (the reason for the ! in the name), so make a copy first if necessary.

NRRD.axes2space

NRRD.axisinfo

NRRD.checked_size

NRRD.chksize

NRRD.colorant_eltype — Function.

colorant_eltype(C, T) -> Tc

Return a valid "inner" element type Tc for colorant type C. When T != Tc, values must be "converted" before they can be interpreted as type C.

NRRD.copy_not_nan!

NRRD.copy_space!

NRRD.eval

NRRD.fieldorder

NRRD.find_datafile

NRRD.fixedtype — Function.

fixedtype(Traw, header) -> Tu

Attempt to interpret type Traw in terms of FixedPoint numbers. The interpretation depends on whether header has a "sample units" field of the form:

sample units: <colorspace> <whitepoint>

There must be a space between colorspace and whitepoint, and colorspace must be one of "gray", "rgb", "rgba", "hsv", "xyz", or their uppercase variants. (Other than "gray", all of these are supported "kinds" values. "gray" does not typically correspond to an axis, which is why it isn't encoded in "kinds".) The presence of any of these words indicates that the data represent an image rather than some other kind of array.

Any other choices are ignored, as "sample units" can also be an arbitrary string.

Examples:

# conventional uint8 grayscale
sample units: gray 255

# a 14-bit grayscale camera (numbers can be represented in hex format)
sample units: gray 0x3fff

# RGB encoded with float or double
sample units: rgb (1.0,1.0,1.0)

# RGB encoded with float or double, but using the scaling of uint8
sample units: rgb (255.0,255.0,255.0)

# conventional XYZ
sample units: xyz (95.047,100.000,108.883)

# HSV, hue measured in degrees
sample units: hsv (360, 0, 1)

# HSV, hue normalized to [0, 1]
sample units: hsv (1, 0, 1)

If Traw cannot be interpreted as Normed, Tu = Traw.

NRRD.fixedtype_max

NRRD.get_axes

NRRD.get_size

NRRD.header_eltype!

NRRD.headerinfo

NRRD.include

NRRD.isdefaultname

NRRD.load

NRRD.myendian

NRRD.nrrd_format

NRRD.nrrd_parse

NRRD.nrrd_write

NRRD.nrrd_write_elty

NRRD.numberparse

NRRD.outer_eltype! — Function.

outer_eltype!(header, Traw) -> T, nd, perm

Extract the julia array eltype T, the number of dimensions nd excluding color/complex/vector/matrix element data, and any permutation needed to put the eltype dimension first. Any dimensions in the header corresponding to color (or if "kind" is set to one of the vector types) will be "consumed" upon exit. Traw is the element type as determined by raw_eltype.

See also: raw_eltype.

NRRD.parse_header — Function.

parse_header(io) -> version, header, keyvals, comments

Parse the NRRD header (the top of the .nrrd or the separate .nhdr file). io should be positioned just after the initial "NRRD" in the file. This reads up to and including the first blank line of the file, so at the end (if this is a properly-formatted NRRD file) io is positioned at the first byte of the data (if present).

Outputs:

version is a 4-character string, e.g., "0002", giving the NRRD version of the header.

header is a Dict{String,Any} of field=>setting pairs (the settings are parsed, they are not necessarily strings)

keyvals is a Dict{String,String} containing key=>value pairs (NRRD0002 or higher, lines like key:=value; many NRRD files do not contain any of these)

comments is an array containing lines of the header that began with # (but with the # and leading whitespace stripped out)

See also: write_header.

NRRD.parse_type

NRRD.per_axis

NRRD.per_spacedim

NRRD.raw_eltype — Function.

raw_eltype(header) -> Traw, need_bswap
raw_eltype(::Type{T}) -> Traw

Get the "basic" element type of the data, e.g., UInt16 or Float32.

This function does not try to determine whether the image is color (Traw does not contain any color information), nor does it try to interpret Traw as a Normed type.

See also: outer_eltype!, fixedtype.

NRRD.safelength

NRRD.save

NRRD.space2axes

NRRD.spacings

NRRD.startsteplen

NRRD.startstepstop

NRRD.startstoplen

NRRD.startstopsteplen

NRRD.stream2name

NRRD.string2type

NRRD.type2string

NRRD.unit_string_dict

NRRD.version

NRRD.version_reqs

NRRD.write_header — Function.

write_header(io, version, header, [keyvals, [comments]])

Write an NRRD header, as the top of the .nrrd or the separate .nhdr file. io should be positioned just after the initial "NRRD" in the file. This writes the header and a blank line, so that at the end io is positioned at the first byte of the data (if present).

Note that if you're writing a header for a "detached" data file (separate .nhdr and .raw files), header should contain a "data file" or "datafile" field storing the name of the .raw file.

Inputs:

version is a 4-character string, e.g., "0002", giving the NRRD version of the header, or a integer corresponding to a recognized NRRD header version number

header is a Dict{String,Any} of field=>setting pairs (as returned by parse_header)

keyvals is a Dict{String,String} containing key=>value pairs (NRRD0002 or higher, lines like key:=value; many NRRD files do not contain any of these)

comments is an array containing lines of the header that began with # (but with the # and leading whitespace stripped out)

See also: parse_header.

NRRD.writeversionstr