Docstrings

SLEEF.Double

SLEEF.FMA_FAST

SLEEF.FP_ILOGB0

SLEEF.FP_ILOGBNAN

SLEEF.INT_MAX

SLEEF.L10L

SLEEF.L10U

SLEEF.L2L

SLEEF.L2U

SLEEF.M1SQRT2

SLEEF.M2P13

SLEEF.M2P23

SLEEF.MD2P13

SLEEF.MD2P23

SLEEF.MDLN10E

SLEEF.MDLN2

SLEEF.MDLN2E

SLEEF.MDPI

SLEEF.MDPI2

SLEEF.MLN2

SLEEF.MLN2E

SLEEF.MLOG10_2

SLEEF.MSQRT2

SLEEF.M_1_PI

SLEEF.M_2_PI

SLEEF.M_2_PI_H

SLEEF.M_2_PI_L

SLEEF.M_4_PI

SLEEF.M_PI

SLEEF.PI_2

SLEEF.PI_4

SLEEF.PI_A

SLEEF.PI_B

SLEEF.PI_C

SLEEF.PI_D

SLEEF.PI_XD

SLEEF.PI_XE

SLEEF.SLEEF

SLEEF.SQRT_MAX

SLEEF.TRIG_MAX

SLEEF._sign

SLEEF._split_exponent — Function.

A helper function for ldexpk

First note that r = (q >> n) << n clears the lowest n bits of q, i.e. returns 2^n where n is the largest integer such that q >= 2^n

For numbers q less than 2^m the following code does the same as the above snippet r = ( (q>>v + q) >> n - q>>v ) << n For numbers larger than or equal to 2^v this subtracts 2^n from q for q>>n times.

The function returns q(input) := q(output) + offset*r

In the code for ldexpk we actually use m = ( (m>>n + m) >> n - m>>m ) << (n-2). So that x has to be multplied by u four times x = x*u*u*u*u to put the value of the offset exponent amount back in.

SLEEF._sqrt

SLEEF.acos — Function.

acos(x)

Compute the inverse cosine of x, where the output is in radians.

SLEEF.acos_fast — Function.

acos_fast(x)

Compute the inverse cosine of x, where the output is in radians.

SLEEF.acosh — Function.

acosh(x)

Compute the inverse hyperbolic cosine of x.

SLEEF.asin — Function.

asin(x)

Compute the inverse sine of x, where the output is in radians.

SLEEF.asin_fast — Function.

asin_fast(x)

Compute the inverse sine of x, where the output is in radians.

SLEEF.asinh — Function.

asinh(x)

Compute the inverse hyperbolic sine of x.

SLEEF.atan — Function.

atan(x)

Compute the inverse tangent of x, where the output is in radians.

atan(x, y)

Compute the inverse tangent of x/y, using the signs of both x and y to determine the quadrant of the return value.

SLEEF.atan2_fast

SLEEF.atan2k

SLEEF.atan2k_fast

SLEEF.atan2k_fast_kernel

SLEEF.atan2k_kernel

SLEEF.atan_fast — Function.

atan_fast(x)

Compute the inverse tangent of x, where the output is in radians.

atan2_fast(x, y)

Compute the inverse tangent of x/y, using the signs of both x and y to determine the quadrant of the return value.

SLEEF.atan_fast_kernel

SLEEF.atanh — Function.

atanh(x)

Compute the inverse hyperbolic tangent of x.

SLEEF.cbrt

SLEEF.cbrt_fast

SLEEF.cbrt_kernel

SLEEF.cos — Function.

cos(x)

Compute the cosine of x, where the output is in radians.

SLEEF.cos_fast — Function.

cos_fast(x)

Compute the cosine of x, where the output is in radians.

SLEEF.cosh — Function.

cosh(x)

Compute hyperbolic cosine of x.

SLEEF.dadd

SLEEF.dadd2

SLEEF.ddiv

SLEEF.dmul

SLEEF.dnormalize

SLEEF.drec

SLEEF.dsqrt

SLEEF.dsqu

SLEEF.dsub

SLEEF.dsub2

SLEEF.eval

SLEEF.exp — Function.

exp(x)

Compute the base-e exponential of x, that is eˣ.

SLEEF.exp10 — Function.

exp10(x)

Compute the base-10 exponential of x, that is 10ˣ.

SLEEF.exp10_kernel

SLEEF.exp2 — Function.

exp2(x)

Compute the base-2 exponential of x, that is 2ˣ.

SLEEF.exp2_kernel

SLEEF.exp_kernel

SLEEF.expk

SLEEF.expk2

SLEEF.expk2_kernel

SLEEF.expk_kernel

SLEEF.expm1 — Function.

expm1(x)

Compute eˣ- 1 accurately for small values of x.

SLEEF.float2integer

SLEEF.hypot — Function.

hypot(x,y)

Compute the hypotenuse \sqrt{x^2+y^2} avoiding overflow and underflow.

SLEEF.ilogb — Function.

ilogb(x)

Returns the integral part of the logarithm of abs(x), using base 2 for the logarithm. In other words, this computes the binary exponent of x such that

x = significand × 2^exponent,

where significand ∈ [1, 2).

Exceptional cases (where Int is the machine wordsize)

x = 0 returns FP_ILOGB0

x = ±Inf returns INT_MAX

x = NaN returns FP_ILOGBNAN

SLEEF.ilogb2k

SLEEF.ilogbk — Function.

ilogbk(x) -> Int

Returns the integral part of the logarithm of |x|, using 2 as base for the logarithm; in other words this returns the binary exponent of x so that

x = significand × 2^exponent

where significand ∈ [1, 2).

SLEEF.include

SLEEF.integer2float

SLEEF.is_fma_fast

SLEEF.isnegzero

SLEEF.isninf

SLEEF.ispinf

SLEEF.ldexp — Function.

ldexp(a, n)

Computes a × 2^n

SLEEF.ldexp2k

SLEEF.ldexp3k

SLEEF.ldexpk — Function.

ldexpk(a, n)

Computes a × 2^n.

SLEEF.log — Function.

log(x)

Compute the natural logarithm of x. The inverse of the natural logarithm is the natural expoenential function exp(x)

SLEEF.log10 — Function.

log10(x)

Returns the base 10 logarithm of x.

SLEEF.log1p — Function.

log1p(x)

Accurately compute the natural logarithm of 1+x.

SLEEF.log2 — Function.

log2(x)

Returns the base 2 logarithm of x.

SLEEF.log_fast — Function.

log_fast(x)

Compute the natural logarithm of x. The inverse of the natural logarithm is the natural expoenential function exp(x)

SLEEF.log_fast_kernel

SLEEF.log_kernel

SLEEF.logk

SLEEF.logk2

SLEEF.logk2_kernel

SLEEF.logk_kernel

SLEEF.max_exp

SLEEF.max_exp10

SLEEF.max_exp2

SLEEF.max_expm1

SLEEF.min_exp

SLEEF.min_exp10

SLEEF.min_exp2

SLEEF.min_expm1

SLEEF.over_log1p

SLEEF.over_sch

SLEEF.over_th

SLEEF.pow — Function.

pow(x, y)

Exponentiation operator, returns x raised to the power y.

SLEEF.pow2i

SLEEF.scale

SLEEF.sin — Function.

sin(x)

Compute the sine of x, where the output is in radians.

SLEEF.sin_fast — Function.

sin_fast(x)

Compute the sine of x, where the output is in radians.

SLEEF.sincos — Function.

sincos(x)

Compute the sin and cosine of x simultaneously, where the output is in radians, returning a tuple.

SLEEF.sincos_a_kernel

SLEEF.sincos_b_kernel

SLEEF.sincos_fast — Function.

sincos_fast(x)

Compute the sin and cosine of x simultaneously, where the output is in radians, returning a tuple.

SLEEF.sincos_fast_kernel

SLEEF.sincos_kernel

SLEEF.sinh — Function.

sinh(x)

Compute hyperbolic sine of x.

SLEEF.split_exponent

SLEEF.splitprec

SLEEF.tan — Function.

tan(x)

Compute the tangent of x, where the output is in radians.

SLEEF.tan_fast — Function.

tan_fast(x)

Compute the tangent of x, where the output is in radians.

SLEEF.tan_fast_kernel

SLEEF.tan_kernel

SLEEF.tanh — Function.

tanh(x)

Compute hyperbolic tangent of x.

SLEEF.threshold_exponent

SLEEF.trunclo

SLEEF.under_atan2

SLEEF.under_expk