Docstrings

PooledArrays.DEFAULT_POOLED_REF_TYPE

PooledArrays.PooledArray — Type.

PooledArray(array, [reftype])

Convert the given array to a PooledArray where each element will be referenced as an integer of the given type. If no reftype is specified one is chosen automatically based on the number of unique elements.

PooledArrays.PooledArrays

PooledArrays.PooledMatrix

PooledArrays.PooledVector

PooledArrays.RefArray

PooledArrays._invert

PooledArrays._label

PooledArrays._perm

PooledArrays._vcat!

PooledArrays._widen

PooledArrays.eval

PooledArrays.fast_sortable

PooledArrays.getpoolidx

PooledArrays.groupsort_indexer

PooledArrays.include

PooledArrays.unsafe_pool_push!