Docstrings

Espresso.@get — Macro.

Same as get function, but evaluates default_expr only if needed

Espresso.@get_array — Macro.

Get array of size sz from a dict by key. If element doesn't exist or its size is not equal to sz, create and return new array using default_expr. If element exists, but is not an error, throw ArgumentError.

Espresso.@get_or_create — Macro.

Same as @get, but creates new object from default_expr if it didn't exist before

Espresso.@get_or_return — Macro.

Same as @get, but immediately exits function and return default_expr if key doesn't exist.

Espresso.@inplacerule

Espresso.@simple_rule — Macro.

Macro to add simplification rules. Example:

@simple_rule (-x * -y) (x * y)

where (-x * -y) is a pattern to match expression and (x * y) is what it should be transformed to (see rewrite() to understand expression rewriting). Symbols Set(Symbol[:a, :b, :y, :x]) may be used as placeholders when defining new rules, all other symbols will be taken literally.

Espresso.@tracked — Macro.

Define a function or broadcasting rule for the specified signature which computes the result as for ordinary (not tracked) data and writes it to the graph.

Note: this function expects at least 1 parameter of TrackedReal or TrackedArray type with name x.

Espresso.AbstractExGraph

Espresso.BufCodeGen

Espresso.CONST_OPS

Espresso.CUDA_NATIVE_RULES

Espresso.CuCodeGen

Espresso.CuVecCodeGen

Espresso.DEFAULT_GRAPH

Espresso.DEFAULT_PHS

Espresso.DOT_OPS

Espresso.Espresso

Espresso.ExGraph

Espresso.ExH

Espresso.ExNode

Espresso.FT

Espresso.GPUCodeGen

Espresso.INPLACE_PHS

Espresso.INPLACE_RULES

Espresso.KNOWN_MODULES

Espresso.Mat

Espresso.Num

Espresso.Numeric

Espresso.OLD_BCAST_OPS

Espresso.OPT_PHS

Espresso.OPT_VEC_RULES

Espresso.OpName

Espresso.PREPROCESS_RULES

Espresso.RECOVER_LOWERED_RULES

Espresso.SIMPLE_PHS

Espresso.SIMPLE_RULES

Espresso.SIMPLE_TO_DOT

Espresso.Symbolic

Espresso.TrackedArray

Espresso.TrackedReal

Espresso.Vec

Espresso.VectorCodeGen

Espresso.__construct

Espresso.__construct_immutable

Espresso._simplify

Espresso.add_inplace_rule

Espresso.add_renaming!

Espresso.arg_names

Espresso.arg_types

Espresso.assign_chain — Function.

Collect all replacable variables from a chain of assignments in a graph. Variables y and x are considered replacable if there's a node y = x and both variables have the same set of guards. Note that this allows nodes to have different sets of indices.

Espresso.assign_chain!

Espresso.assign_chain_index_replacements

Espresso.autoselect_codegen

Espresso.buffer_expr

Espresso.canonical — Function.

Return canonical representation of a function name, e.g.:

Base.+  ==> +
Main.+  ==> + (resolved to Base.+)
Base.LinAlg.exp ==> exp
Mod.foo ==> Mod.foo

Espresso.canonical_calls

Espresso.cat

Espresso.collect_deps

Espresso.collect_deps!

Espresso.concretise_types

Espresso.convert_call

Espresso.cos

Espresso.cuda_buffer_expr

Espresso.dependencies — Function.

Get names of dependenices of this node

Espresso.dependents — Function.

For each variable in graph, calculate all variables that depend on it. This is essentially the opposite of dependencies(nd::ExNode), but operates on variable names rather than nodes.

Espresso.depwarn_eingraph

Espresso.destruct — Function.

Replace all struct arguments by a list of their plain analogues. Example:

args = [:m, :x, :y]
types = (Linear, Matrix{Float64}, Matrix{Float64})
ex = :(sum((m.W * x .+ m.b) - y))

destruct(args, types, ex)
# ==>
# ([:m_W, :m_b, :x, :y],
#  :(sum((m_W * x .+ m_b) - y)),
#  Dict(:(m.W) => :m_W, :(m.b) => :m_b))

Espresso.destruct_inputs

Espresso.dot_expr — Function.

Given a list of symbols such as [:x, :y, :z] constructs expression x.y.z. This is useful for building expressions of qualified names such as Base.LinAlg.exp.

Espresso.eliminate_common

Espresso.eval

Espresso.eval_codegen — Function.

For buffered codegens, return unbuffered version that can be used in evaluate!()

Espresso.eval_tracked!

Espresso.evaluate! — Function.

Evaluate node, i.e. fill its val by evaluating node's expression using values of its dependencies.

Espresso.expand_const — Function.

Expand all constant vars in a given expression

Espresso.expand_deps

Espresso.expand_deps!

Espresso.expand_fixed_sequences — Function.

Look at each node's dependencies and, if there are known pattern sequences, rewrite them in a more optimal way.

Espresso.expr_like

Espresso.external_vars

Espresso.field_values

Espresso.find_indices

Espresso.find_key

Espresso.find_var_names

Espresso.find_vars — Function.

find_vars(ex; rec=true)

Same as get_vars(), but recursive by default

Espresso.findex — Function.

Find sub-expressions matching a pattern. Example:

ex = :(a * f(x) + b * f(y))
pat = :(f(_))
findex(pat, ex)   # ==> [:(f(x)), :(f(y))]

Espresso.findex!

Espresso.flatten — Function.

Flattenx vector of vectors

Espresso.flatten! — Function.

Flatten vector of vectors in-place

Espresso.flatten1 — Function.

Flatten one level of nested vectors

Espresso.flatten_dots — Function.

Given a list of expression arguments, flatten the dotted ones. Example:

args = [:foo, :([a, b, c]...)]
flatten_dots(args)
# ==> [:foo, :a, :b, :c]

Espresso.flip

Espresso.forall_indices

Espresso.forall_sum_indices

Espresso.force_bitness

Espresso.func_expr

Espresso.func_mod

Espresso.func_name

Espresso.funexpr

Espresso.fuse_assigned — Function.

Collapse unnecessary assignment nodes, rewriting all affected nodes. Example:

tmp1 = x * y
z = tmp1

will be rewritten to

z = x * y

Espresso.fuse_broadcasting

Espresso.fuse_broadcasting_node

Espresso.generate_code

Espresso.genname — Function.

Generate a new unique name for intermediate variable in graph

Espresso.gennames

Espresso.get_arg_names

Espresso.get_arg_types

Espresso.get_caller_module

Espresso.get_default_graph

Espresso.get_indices

Espresso.get_method

Espresso.get_or_generate_argnames

Espresso.get_source_at — Function.

Looks up the source of method in the file path found in method. Returns the AST and source string, might throw an LoadError if file not found.

Espresso.get_source_file

Espresso.get_var_names

Espresso.get_vars — Function.

Get variables (Symbol or Expr(:ref)) involved in exprssion

Espresso.get_vars!

Espresso.getcategory

Espresso.getexpr

Espresso.getexpr_kw

Espresso.getguards

Espresso.getsize

Espresso.getvalue

Espresso.getvar

Espresso.gpu_buffer_expr

Espresso.graph_hash — Function.

A single number to represent a graph. Insensitive to variable names.

Espresso.inc_var_name — Function.

Given a symbolic name, either adds 2 to the end or increment existing number. Example:

inc_var_name(:x)   # ==> x2
inc_var_name(:x2)  # ==> x3

Espresso.include

Espresso.indexof

Espresso.inline_nodes

Espresso.inline_subgraphs

Espresso.is_bcast — Function.

Check if all operations in this expression are broadcasting

Espresso.is_bcast_vec

Espresso.is_calculable

Espresso.iscall

Espresso.isconv

Espresso.isindexed

Espresso.isparameters

Espresso.isplaceholder

Espresso.isref

Espresso.isstruct — Function.

Check if an object is of a struct type, i.e. not a number or array

Espresso.istemp

Espresso.istracked

Espresso.jlhome

Espresso.juliabasepath

Espresso.make_call_expr — Function.

Make call expression from function name, ordinary and keyword arguments.

The reverse of this operation is parsecallexpr()

Espresso.make_elementwise

Espresso.make_func_expr

Espresso.make_indexed — Function.

Make indexed variable. Examples:

make_indexed(:x, [])       ==> :x
make_indexed(:x, [:i,:j])  ==> :(x[i,j])

See also: split_indexed

Espresso.make_kw_params

Espresso.make_subgraph — Function.

Find definition of a called function and build its subgraph ready for inlining

Espresso.matchex — Function.

Match expression ex to a pattern pat, return nullable dictionary of matched symbols or rpatpressions. Example:

ex = :(u ^ v)
pat = :(_x ^ _n)
matchex(pat, ex)
# ==> Union{ Dict{Symbol,Any}(:_n=>:v,:_x=>:u), Void }

NOTE: two symbols match if they are equal or symbol in pattern is a placeholder. Placeholder is any symbol that starts with ''. It's also possible to pass list of placeholder names (not necessarily starting wiht '') via phs parameter:

ex = :(u ^ v)
pat = :(x ^ n)
matchex(pat, ex; phs=Set([:x, :n]))
# ==> Union{ Dict{Symbol,Any}(:n=>:v,:x=>:u), Void } 

Several elements may be matched using ... expression, e.g.:

ex = :(A[i, j, k])
pat = :(x[I...])
matchex(pat, ex; phs=Set([:x, :I]))
# ==> Union{ Dict(:x=>:A, :I=>[:i,:j,:k]), Void }

Optional parameters:

phs::Set{Symbol} = DEFAULT_PHS[1] A set of placeholder symbols

allow_ex::Boolean = true Allow matchinng of symbol pattern to an expression. Example:

    matchex(:(_x + 1), :(a*b + 1); allow_ex=true)  # ==> matches
    matchex(:(_x + 1), :(a*b + 1); allow_ex=false)  # ==> doesn't match

exact::Boolean = false Allow matching of the same expression to different keys

    matchex(:(_x + _y), :(a + a); exact=false) # ==> matches
    matchex(:(_x = _y), :(a + a); exact=true)  # ==> doesn't match

Espresso.matchex!

Espresso.matchingex — Function.

Check if expression matches pattern. See matchex() for details.

Espresso.mergeex

Espresso.mk_eval_expr

Espresso.mul!

Espresso.named_field_values

Espresso.parse! — Function.

Parse Julia expression and build ExGraph in-place. Return the the output variable.

Espresso.parse_call_args — Function.

Given a call expression, parse regular and keyword arguments

See also: split_params

Espresso.parse_call_expr — Function.

Parse call expression into function name, ordinary and keyword arguments. :kw and :parameters arguments are treated the same way.

The reverse of this operation is makecallexpr()

Espresso.preprocess

Espresso.prop_subs — Function.

Propagate substitution rules. Example:

Dict(
    :x => y,
    :y => z
)

is transformed into:

Dict(
    :x => z,
    :y => z
)

Espresso.recover_lowered — Function.

Try to recover an expression from a lowered form. Example:

ex = (Main.sum)((Base.literal_pow)(Main.^, (Base.broadcast)(Main.-, (Main.predict)(W, b, x), y), (Core.apply_type)(Base.Val, 2)))

Espresso.recover_lowered_rec

Espresso.reindex_from_beginning

Espresso.remember_size!

Espresso.remove_unused — Function.

Removes unused variables from multiline expressions, e.g. in:

x = u * v
y = x + 1
z = 2x

y isn't used to compute output variable z, so it's removed:

x = u * v
z = 2x

Espresso.rename

Espresso.rename!

Espresso.rename_from_beginning

Espresso.rename_repeated

Espresso.reparse

Espresso.replace_inner_constr — Function.

Replace all calls to an inner constructor with the corresponding outer constructor

Espresso.replace_slots

Espresso.reset_default_graph!

Espresso.reset_tape

Espresso.rewrite — Function.

rewrite(ex, pat, rpat)

Rewrite expression ex according to a transform from pattern pat to a substituting expression rpat. Example (derivative of x^n):

ex = :(u ^ v)
pat = :(_x ^ _n)
rpat = :(_n * _x ^ (_n - 1))
rewrite(ex, pat, rpat) # ==> :(v * u ^ (v - 1))

Espresso.rewrite_all — Function.

rewrite_all(ex, rules)

Recursively rewrite an expression according to a list of rules like [pat => rpat] Example:

ex = :(foo(bar(foo(A))))
rules = [:(foo(x)) => :(quux(x)),
         :(bar(x)) => :(baz(x))]
rewrite_all(ex, rules; phs=[:x])
# ==> :(quux(baz(quux(A))))

rewrite_all(ex, pat, rpat)

Recursively rewrite all occurrences of a pattern in an expression. Example:

ex = :(foo(bar(foo(A))))
pat = :(foo(x))
rpat = :(quux(x))
rewrite_all(ex, pat, rpat; phs=[:x])
# ==> :(quux(bar(quux(A))))

Espresso.sanitize

Espresso.set_default_graph

Espresso.set_default_placeholders

Espresso.setexpr!

Espresso.setexpr_kw!

Espresso.setguards!

Espresso.setvalue!

Espresso.setvar!

Espresso.simplify — Function.

Simplify expression x by applying a set of rules. Common examples of simplification include calculation of fully numeric subexpressions, removing needless multiplication by 1, etc.

Use macro @simple_rule to add new simplification rules.

Espresso.sin

Espresso.split_indexed — Function.

Split possibly indexed variable into a name and indices. Examples:

split_indexed(:x)         ==> (:x, [])
split_indexed(:(x[i,j]))  ==> (:x, [:i,:j])

See also: make_indexed

Espresso.split_params — Function.

Split parameters of a function signature, returning a list of (param name, param type) tuples and a list of keyword parameters.

See also: parsecallargs

Espresso.squeeze_sum

Espresso.squeeze_sum_1

Espresso.squeeze_sum_2

Espresso.struct_like

Espresso.subgraph_interm_subs_table

Espresso.subs — Function.

Substitute symbols in ex according to substitute table st. Example:

ex = :(x ^ n)
subs(ex, x=2)            # ==> :(2 ^ n)

alternatively:

subs(ex, Dict(:x => 2))  # ==> :(2 ^ n)

If ex contains a :(xs...) argument and st contains an array-valued sabstitute for it, the substitute will be flattened:

ex = :(foo(xs...))
subs(ex, Dict(:xs => [:a, :b, :c]))
# ==> :(foo(a, b, c))

Espresso.subs_bcast_with_dot

Espresso.sum_1

Espresso.sum_2

Espresso.sum_indices

Espresso.swap_default_graph!

Espresso.to_block

Espresso.to_buffered

Espresso.to_context

Espresso.to_dict

Espresso.to_einsum_expr — Function.

Convert ExNode to a fortmat compatible with Einsum.jl

Espresso.to_expr — Function.

Convert ExNode to a full expression, e.g. for vectorized notation:

z = x + y

or for indexed notation:

z[i] = x[i] + y[i]

Espresso.to_expr_kw — Function.

Same as to_expr(ExNode), but includes keyword arguments if any

Espresso.to_inplace

Espresso.topsort — Function.

Sort graph topologically

Espresso.topsort_visit!

Espresso.track_bcast

Espresso.track_call

Espresso.tracked

Espresso.tracked_exgraph

Espresso.tracked_val

Espresso.tryeval

Espresso.tryoptimize

Espresso.tryrewrite — Function.

Same as rewrite, but returns Union{Expr, Void} and doesn't throw an error when expression doesn't match pattern

Espresso.unzip

Espresso.value

Espresso.varidxs

Espresso.varname

Espresso.with_guards

Espresso.with_indices — Function.

Generate index names and make an indexed variable using them

Espresso.with_keywords

Espresso.without — Function.

Remove rpatpression conforming to a pattern. Example:

ex = :(x * (m == n))
pat = :(_i == _j)
ex = without(ex, pat)  # ==> :x

Espresso.without_keywords — Function.

Remove all :kw and :parameters nodes from a call expression

Espresso.without_types