Docstrings

LanguageServer.@json_read

LanguageServer.ApplyWorkspaceEditParams

LanguageServer.ApplyWorkspaceEditResponse

LanguageServer.CancelParams

LanguageServer.Capabilities

LanguageServer.ClientCapabilities

LanguageServer.CodeActionCapabilities

LanguageServer.CodeActionContext

LanguageServer.CodeActionKindCapabilities

LanguageServer.CodeActionLiteralCapabilities

LanguageServer.CodeActionParams

LanguageServer.CodeLens

LanguageServer.CodeLensOptions

LanguageServer.CodeLensParams

LanguageServer.CodeLensRegistrationOptions

LanguageServer.Command

LanguageServer.CompletionCapabilities

LanguageServer.CompletionContext

LanguageServer.CompletionItem

LanguageServer.CompletionItemCapabilities

LanguageServer.CompletionItemKind

LanguageServer.CompletionItemKindCapabilities

LanguageServer.CompletionList

LanguageServer.CompletionOptions

LanguageServer.CompletionParams

LanguageServer.CompletionTriggerKind

LanguageServer.ConfigurationItem

LanguageServer.ConfigurationParams

LanguageServer.DefaultTypeConstructorLoc

LanguageServer.Diagnostic

LanguageServer.DiagnosticRelatedInformation

LanguageServer.DiagnosticSeverity

LanguageServer.DidChangeConfiguration

LanguageServer.DidChangeTextDocumentParams

LanguageServer.DidChangeWatchedFilesParams

LanguageServer.DidChangeWatchedFilesRegistrationOptions

LanguageServer.DidCloseTextDocumentParams

LanguageServer.DidOpenTextDocumentParams

LanguageServer.DidSaveTextDocumentParams

LanguageServer.Document

LanguageServer.DocumentFilter

LanguageServer.DocumentFormattingParams

LanguageServer.DocumentHighlight

LanguageServer.DocumentHighlightKind

LanguageServer.DocumentLink

LanguageServer.DocumentLinkOptions

LanguageServer.DocumentLinkParams

LanguageServer.DocumentOnTypeFormattingOptions

LanguageServer.DocumentOnTypeFormattingParams

LanguageServer.DocumentOnTypeFormattingRegistrationOptions

LanguageServer.DocumentRangeFormattingParams

LanguageServer.DocumentSelector

LanguageServer.DocumentSymbolCapabilities

LanguageServer.DocumentSymbolParams

LanguageServer.DocumentUri

LanguageServer.ExecuteCommandOptions

LanguageServer.ExecuteCommandParams

LanguageServer.ExecuteCommandRegistrationOptions

LanguageServer.FileChangeType

LanguageServer.FileChangeType_Changed

LanguageServer.FileChangeType_Created

LanguageServer.FileChangeType_Deleted

LanguageServer.FileEvent

LanguageServer.FileSystemWatcher

LanguageServer.FormattingOptions

LanguageServer.Hover

LanguageServer.HoverCapabilities

LanguageServer.InitializeParams

LanguageServer.InitializeResult

LanguageServer.Invoked

LanguageServer.JSONRPC

LanguageServer.LSDiagnostic

LanguageServer.LanguageServer

LanguageServer.LanguageServerInstance

LanguageServer.Location

LanguageServer.LogMessageParams

LanguageServer.MarkedString

LanguageServer.MarkupContent

LanguageServer.Message

LanguageServer.MessageActionItem

LanguageServer.MessageType

LanguageServer.ParameterInformation

LanguageServer.Position

LanguageServer.PublishDiagnosticsCapabilities

LanguageServer.PublishDiagnosticsParams

LanguageServer.Range

LanguageServer.ReferenceContext

LanguageServer.ReferenceParams

LanguageServer.Registration

LanguageServer.RegistrationParams

LanguageServer.RenameParams

LanguageServer.SaveOptions

LanguageServer.ServerCapabilities

LanguageServer.ShowMessageParams

LanguageServer.ShowMessageRequestParams

LanguageServer.SignatureCapabilities

LanguageServer.SignatureHelp

LanguageServer.SignatureHelpOptions

LanguageServer.SignatureHelpRegistrationOptions

LanguageServer.SignatureInformation

LanguageServer.SignatureInformationCapabilities

LanguageServer.StaticRegistrationOptions

LanguageServer.SymbolCapabilities

LanguageServer.SymbolInformation

LanguageServer.SymbolKind

LanguageServer.SymbolKindCapabilities

LanguageServer.SynchronizationCapabilities

LanguageServer.T

LanguageServer.TextDocumentChangeRegistrationOptions

LanguageServer.TextDocumentClientCapabilities

LanguageServer.TextDocumentContentChangeEvent

LanguageServer.TextDocumentEdit

LanguageServer.TextDocumentIdentifier

LanguageServer.TextDocumentItem

LanguageServer.TextDocumentPositionParams

LanguageServer.TextDocumentReason

LanguageServer.TextDocumentRegistrationOptions

LanguageServer.TextDocumentSyncKind

LanguageServer.TextDocumentSyncOptions

LanguageServer.TextEdit

LanguageServer.TriggerCharacter

LanguageServer.TriggerForIncompleteCompletion

LanguageServer.URI2 — Type.

A type that represents a URI. The main feature provided currently is that when the file:// scheme is used and the platform is Windows, equality and hashing is case-insensitive.

LanguageServer.Unregistration

LanguageServer.UnregistrationParams

LanguageServer.VersionedTextDocumentIdentifier

LanguageServer.WatchKind

LanguageServer.WillSaveTextDocumentParams

LanguageServer.WorkspaceClientCapabilities

LanguageServer.WorkspaceEdit

LanguageServer.WorkspaceEditCapabilities

LanguageServer.WorkspaceFolder

LanguageServer.WorkspaceFoldersChangeEvent

LanguageServer.WorkspaceFoldersOptions

LanguageServer.WorkspaceOptions

LanguageServer.WorkspaceSymbolParams

LanguageServer._offset_unitrange

LanguageServer.clear_diagnostics

LanguageServer.convert_diagnostic

LanguageServer.didChangeWorkspaceFoldersParams

LanguageServer.eval

LanguageServer.filepath2uri

LanguageServer.find_ref

LanguageServer.find_references

LanguageServer.find_root

LanguageServer.get_line_offsets

LanguageServer.get_locations

LanguageServer.get_offset

LanguageServer.get_open_in_editor

LanguageServer.get_partial_completion

LanguageServer.get_position_at

LanguageServer.get_sig_args

LanguageServer.get_signatures

LanguageServer.get_text

LanguageServer.get_toks

LanguageServer.haskeynotnull

LanguageServer.hasreadperm

LanguageServer.include

LanguageServer.is_ignored

LanguageServer.is_workspace_file

LanguageServer.isjuliabasedir

LanguageServer.iter_lines

LanguageServer.joinuriwithpath

LanguageServer.kw_completion

LanguageServer.latex_completions

LanguageServer.load_folder

LanguageServer.load_rootpath

LanguageServer.parse_all

LanguageServer.parse_jmd

LanguageServer.process

LanguageServer.publish_diagnostics

LanguageServer.read_transport_layer

LanguageServer.remove_workspace_files

LanguageServer.send

LanguageServer.serverCapabilities

LanguageServer.serverbusy

LanguageServer.serverready

LanguageServer.set_open_in_editor

LanguageServer.should_file_be_linted

LanguageServer.str_value

LanguageServer.toggle_file_lint

LanguageServer.update

LanguageServer.uri2filepath

LanguageServer.write_transport_layer