Docstrings

KernelDensityEstimate.*

KernelDensityEstimate.BallTree

KernelDensityEstimate.BallTreeDensity — Type.

Evaluate the KDE object at given points.

Note, must use Array{Float64,2} when passing in evaluation points.

KernelDensityEstimate.DirectSize

KernelDensityEstimate.FORCE_EVAL_DIRECT

KernelDensityEstimate.GaussianKer

KernelDensityEstimate.GbGlb

KernelDensityEstimate.KernelDensityEstimate

KernelDensityEstimate.MSCompOpt

KernelDensityEstimate.MixtureDensity

KernelDensityEstimate.NO_CHILD

KernelDensityEstimate.Ndim

KernelDensityEstimate.Npts

KernelDensityEstimate.VectorRange

KernelDensityEstimate.buildBall!

KernelDensityEstimate.buildTree!

KernelDensityEstimate.bw

KernelDensityEstimate.bwMax

KernelDensityEstimate.bwMin

KernelDensityEstimate.bwUniform

KernelDensityEstimate.calcIndices! — Function.

calcIndices!(glb)

Recompute particles, variance.

KernelDensityEstimate.calcStats!

KernelDensityEstimate.calcStatsBall!

KernelDensityEstimate.calcStatsDensity!

KernelDensityEstimate.center

KernelDensityEstimate.changeWeights!

KernelDensityEstimate.closer!

KernelDensityEstimate.convert

KernelDensityEstimate.dontRecurseSubtrees

KernelDensityEstimate.entropy

KernelDensityEstimate.eval

KernelDensityEstimate.evalAvgLogL

KernelDensityEstimate.evalDirect

KernelDensityEstimate.evaluate

KernelDensityEstimate.evaluateDualTree

KernelDensityEstimate.gaussianProductMeanCov! — Function.

gaussianProductMeanCov!(glb, j, destMu, destCov, idx, skip)
gaussianProductMeanCov!(glb, j, destMu, destCov, idx, skip, addop)
gaussianProductMeanCov!(glb, j, destMu, destCov, idx, skip, addop, getMu)
gaussianProductMeanCov!(glb, j, destMu, destCov, idx, skip, addop, getMu, getLambda)

Multiplication of Gaussians using leave in densities (if skip skip > 0). For on-manifold operations, set getMu and getLambda operations accordingly.

Notes

Used twice in samplePoint! (won't skip) and sampleIndex (will skip LOO).

Assumes manifold diffop baked into getMu.

use jth dimension

KernelDensityEstimate.getBW — Function.

getBW(bd)
getBW(bd, ind)

Return the bandwidths used for each kernel in the density estimate.

KernelDensityEstimate.getEuclidLambda — Function.

getEuclidLambda(lambdas)

Returns Λ=Σ^{-1} as sum of individual information matrices (inverse covariances – i.e. bandwidths).

\[Λ = Σ_i Λ_i\]

KernelDensityEstimate.getEuclidMu — Function.

getEuclidMu(mus, lambdas)
getEuclidMu(mus, lambdas, scale)

Returns Λμ.

\[Λμ = Σ_i (Λ_i*μ_i)\]

KernelDensityEstimate.getIndexOf

KernelDensityEstimate.getKDEMax

KernelDensityEstimate.getKDEMean

KernelDensityEstimate.getKDERange

KernelDensityEstimate.getKDERangeLinspace

KernelDensityEstimate.getKDEfit

KernelDensityEstimate.getMiniMaxi — Function.

getMiniMaxi(bt, leftI, rightI, d, addop, diffop)

Return "smaller" and "larger" of two child nodes.

KernelDensityEstimate.getPoints — Function.

getPoints(bd)

Return the points (centers) used to construct the KDE.

KernelDensityEstimate.getType

KernelDensityEstimate.getWeights — Function.

getWeights(bd)
getWeights(bd, ind)

Return the weights used for each kernel in the density estimate.

KernelDensityEstimate.gibbs1

KernelDensityEstimate.golden — Function.

golden(bd, ax, bx, cx, tol, addop, diffop)

GOLDEN Minimize the nLOO_LL function for KDE bandwidth selection of one variable using golden section search.

xmin, fmin = golden(npd, f, ax, bx, cx, tol) computes a local minimum of f = KDE.nLOO_LL. xmin is the computed local minimizer of f and fmin is f(xmin). xmin is computed to an relative accuracy of TOL.

The parameters ax, bx and cx must satisfy the following conditions: ax < bx < cx, f(bx) < f(ax) and f(bx) < f(cx).

xmin satisfies ax < xmin < cx. golden is guaranteed to succeed if f is continuous between ax and cx

Roman Geus, ETH Zuerich, 9.12.97

KernelDensityEstimate.include

KernelDensityEstimate.initIndices!

KernelDensityEstimate.intersIntgAppxIS

KernelDensityEstimate.isLeaf

KernelDensityEstimate.kde! — Function.

kde!(points)
kde!(points, ks)
kde!(points, ks, addop)
kde!(points, ks, addop, diffop)

Construct a BallTreeDensity object using points for centers and bandwidth ks.

KernelDensityEstimate.kld

KernelDensityEstimate.ksize

KernelDensityEstimate.leafFirst

KernelDensityEstimate.leafLast

KernelDensityEstimate.left

KernelDensityEstimate.levelDown!

KernelDensityEstimate.levelInit!

KernelDensityEstimate.makeBallTree

KernelDensityEstimate.makeBallTreeDensity

KernelDensityEstimate.makeDualTree

KernelDensityEstimate.makeEmptyGbGlb

KernelDensityEstimate.makeFasterSampleIndex! — Function.

makeFasterSampleIndex!(j, cmo, glb)
makeFasterSampleIndex!(j, cmo, glb, diffop)

Calculate the likelihoods of left out kernel means (label_i's) against Normal distribution with mean glb.Malmost and variance glb.Calmost – we think ...

Assumes

Temporary product gaussian on manifold has already been computed and stored in glb.Calmost and glb.Malmost.

Notes

Incoming temporary product mean is glb.Malmost and covariance glb.Calmost.

j is the left out density – i.e. the density for which we want to select a new label (kernel)

cmo.tmpM is the manifold difference from all dimensions prior to evaluating Gaussian pdf, and relates to levels of the k-d tree...

z is index of leave one out density kernels – i.e. evaluate likelihood of each kernel mean glb.p[z] = p(μ_z)

New kernel will be selected randomly from likelihood weights in glb.p[z:z].

Potential Concerns

Make sure tmpC can happen on Euclidean vs user manifold

KernelDensityEstimate.marginal — Function.

marginal(bd, ind)

Extract the marginal distribution from the given higher dimensional kernel density estimate object.

KernelDensityEstimate.maxDistGauss!

KernelDensityEstimate.maxDistKer!

KernelDensityEstimate.mean

KernelDensityEstimate.minDistGauss!

KernelDensityEstimate.minDistKer!

KernelDensityEstimate.most_spread_coord

KernelDensityEstimate.movePoints!

KernelDensityEstimate.nLOO_LL

KernelDensityEstimate.neighborMinMax

KernelDensityEstimate.pArrHdls

KernelDensityEstimate.parsestringvector

KernelDensityEstimate.printBallTree

KernelDensityEstimate.printGlbs

KernelDensityEstimate.prodAppxMSGibbsS

KernelDensityEstimate.pushDownAll

KernelDensityEstimate.pushDownLocal

Base.rand — Function.

rand(p)
rand(p, N)

Randomly sample points from the KernelDensityEstimate object.

KernelDensityEstimate.randKernel

KernelDensityEstimate.rangeB

KernelDensityEstimate.recurseMinMax

KernelDensityEstimate.recurseOnSubtrees

KernelDensityEstimate.resample

KernelDensityEstimate.right

KernelDensityEstimate.root

StatsBase.sample — Function.

sample(npd, Npts)

Randomly sample points from the KernelDensityEstimate object.

KernelDensityEstimate.sampleIndex — Function.

sampleIndex(j, cmo, glb)
sampleIndex(j, cmo, glb, addop)
sampleIndex(j, cmo, glb, addop, diffop)
sampleIndex(j, cmo, glb, addop, diffop, getMu)
sampleIndex(j, cmo, glb, addop, diffop, getMu, getLambda)

Sample new kernel in leave out density according to multiscale Gibbs sampling.

calculate temporary product of leave in density components (labels previously selected)

evaluate the likelihoods of the kernel means from the left out kernel on temporary product

randomly select a new label (kernel_i) for left out density according to temporarily evaluated likelihoods

Notes

j is the left out density.

Needs on-manifold getMu for product of two leave in density kernels.

Needs diffop (on manifold difference for kernel evaluation).

Sudderth PhD, p.139, Fig. 3.3, top-left operation

KernelDensityEstimate.sampleIndices! — Function.

sampleIndices!(X, cmoi, glb, frm)
sampleIndices!(X, cmoi, glb, frm, diffop)

??

Notes

This function does kernel evaluation internally.

Does not have a loo skip step.

KernelDensityEstimate.samplePoint! — Function.

samplePoint!(X, glb, frm)
samplePoint!(X, glb, frm, addop)
samplePoint!(X, glb, frm, addop, getMu)
samplePoint!(X, glb, frm, addop, getMu, getLambda)

Sampling a point from the product of kernels (density components) listed in glb.variance and glb.particles without skipping a kernel (not a leave-one-out case).

Manifold defined by addop, getMu, and getLambda.

KernelDensityEstimate.select!

KernelDensityEstimate.setForceEvalDirect!

KernelDensityEstimate.string

KernelDensityEstimate.swap!

KernelDensityEstimate.swapBall!

KernelDensityEstimate.swapDensity!

KernelDensityEstimate.updateBandwidth!

KernelDensityEstimate.validIndex

KernelDensityEstimate.variance

KernelDensityEstimate.weight