Docstrings

MAT.@read

MAT.@write

MAT.HDF5_HEADER

MAT.MAT

MAT.MAT_HDF5

MAT.MAT_v5

MAT.eval

HDF5.exists — Function.

exists(matfile_handle, varname) -> Bool

Return true if a variable is present in an opened Matlab file.

See matopen.

MAT.include

MAT.matopen — Function.

matopen(filename [, mode]) -> handle
matopen(f::Function, filename [, mode]) -> f(handle)

Mode defaults to "r" for read. It can also be "w" for write, or "r+" for read or write without creation or truncation.

Use with read, write, close, names, and exists.

MAT.matread — Function.

matread(filename) -> Dict

Return a dictionary of all the variables and values in a Matlab file, opening and closing it automatically.

MAT.matwrite — Function.

matwrite(filename, d::Dict)

Write a dictionary containing variable names as keys and values as values to a Matlab file, opening and closing it automatically.

Base.names — Function.

names(matfile_handle) -> Vector{String}

Return a list of variables in an opened Matlab file.

See matopen.