Docstrings

MLDataPattern.BalancedObs — Type.

BalancedObs([f], data, [count], [obsdim])

Description

Create an iterator that generates count randomly sampled observations from data. In the case count is not provided, it will generate random samples indefinitely.

In contrast to RandomObs, BalancedObs expects data to be a labeled data container. It uses the label distribution of data to make sure every label has an equal probability to be sampled from.

Arguments

f : Optional. A function that should be applied to each observation individually in order to extract or compute the target for that observation. This function is only used once during construction to determine which label each observation belongs to.

data : The object describing the dataset. Can be of any type as long as it implements getobs, nobs, and optionally gettargets (see Details for more information).

count : Optional. The number of randomly sampled observations that the iterator will generate before stopping. If omitted, the iterator will generate randomly sampled observations forever.

obsdim : Optional. If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

Details

For BalancedObs to work on some data structure, the type of the given variable data must implement the labeled data container interface. See ?DataSubset for more info.

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Examples

# load first 55 observations of the iris data as example
# - 50 observations for "setosa"
# -  5 observations for "versicolor"
X, Y = MLDataUtils.load_iris(55)

# go over 100 balanced samples observations in X
num_versicolor = 0
for (x,y) in BalancedObs((X,Y), 100) # also: BalancedObs((X,Y), count = 100)
    @assert typeof(x) <: SubArray{Float64,1}
    @assert length(x) == 4
    # count how many times we sample a versicolor observation
    num_versicolor += y[] == "versicolor" ? 1 : 0
end
println(num_versicolor) # around 50

# if no count it provided the iterator will generate samples forever
for (x,y) in BalancedObs((X,Y))
    # this loop will never stop unless break is used
    if true; break; end
end

see also

RandomObs, targets, RandomBatches, ObsView, BatchView, shuffleobs, DataSubset, BufferGetObs

MLDataPattern.BatchView — Type.

BatchView(data, [size|maxsize], [count], [obsdim])

Description

Create a view of the given data that represents it as a vector of batches. Each batch will contain an equal amount of observations in them. The number of batches and the batch-size can be specified using (keyword) parameters count and size. In the case that the size of the dataset is not dividable by the specified (or inferred) size, the remaining observations will be ignored with a warning.

Note that any data access is delayed until getindex is called, and even getindex returns the result of datasubset which in general avoids data movement until getobs is called.

If used as an iterator, the object will iterate over the dataset once, effectively denoting an epoch. Each iteration will return a mini-batch of constant nobs, which effectively allows to iterator over data one batch at a time.

Arguments

data : The object describing the dataset. Can be of any type as long as it implements getobs and nobs (see Details for more information).

size : The batch-size of each batch. I.e. the number of observations that each batch must contain.

maxsize : The maximum batch-size of each batch. I.e. the number of observations that each batch should contain. If the number of total observation is not divideable by the size it will be reduced until it is.

count : The number of batches that the view will contain.

obsdim : Optional. If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

Methods

Aside from the AbstractArray interface following additional methods are provided.

getobs(data::BatchView, batchindices) : Returns a Vector of the batches specified by batchindices.

nobs(data::BatchView) : Returns the total number of observations in data. Note that unless the batch-size is 1, this number will differ from length.

Details

For BatchView to work on some data structure, the type of the given variable data must implement the data container interface. See ?DataSubset for more info.

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Tom Breloff (Github: https://github.com/tbreloff)

Examples

using MLDataUtils
X, Y = MLDataUtils.load_iris()

A = batchview(X, size = 30)
@assert typeof(A) <: BatchView <: AbstractVector
@assert eltype(A) <: SubArray{Float64,2}
@assert length(A) == 5 # Iris has 150 observations
@assert size(A[1]) == (4,30) # Iris has 4 features

# 5 batches of size 30 observations
for x in batchview(X, size = 30)
    @assert typeof(x) <: SubArray{Float64,2}
    @assert nobs(x) === 30
end

# 7 batches of size 20 observations
# Note that the iris dataset has 150 observations,
# which means that with a batchsize of 20, the last
# 10 observations will be ignored
for (x,y) in batchview((X,Y), size = 20)
    @assert typeof(x) <: SubArray{Float64,2}
    @assert typeof(y) <: SubArray{String,1}
    @assert nobs(x) === nobs(y) === 20
end

# 10 batches of size 15 observations
for (x,y) in batchview((X,Y), maxsize = 20)
    @assert typeof(x) <: SubArray{Float64,2}
    @assert typeof(y) <: SubArray{String,1}
    @assert nobs(x) === nobs(y) === 15
end

# randomly assign observations to one and only one batch.
for (x,y) in batchview(shuffleobs((X,Y)))
    @assert typeof(x) <: SubArray{Float64,2}
    @assert typeof(y) <: SubArray{String,1}
end

# iterate over the first 2 batches of 15 observation each
for (x,y) in batchview((X,Y), size=15, count=2)
    @assert typeof(x) <: SubArray{Float64,2}
    @assert typeof(y) <: SubArray{String,1}
    @assert size(x) == (4, 15)
    @assert size(y) == (15,)
end

see also

eachbatch, ObsView, shuffleobs, getobs, nobs, DataSubset

MLDataPattern.BufferGetObs — Type.

BufferGetObs(iterator, [buffer])

A stateful iterator that stores the output of iterate(iterator,state) into buffer using getobs!(buffer, ...). Depending on the type of data provided by iterator this may be more memory efficient than getobs(...). In the case of array data, for example, this allows for cache-efficient processing of each element without allocating a temporary array.

Note that not all types of data support buffering, because it is the author's choice to opt-in and implement a custom getobs!. For those types that do not provide a custom getobs!, the buffer will be ignored and the result of getobs(...) returned.

see eachobs and eachbatch for concrete examples.

MLDataPattern.DataSubset — Type.

DataSubset(data, [indices], [obsdim])

Description

Used to represent a subset of some data of arbitrary type by storing which observation-indices the subset spans. Furthermore, subsequent subsettings are accumulated without needing to access actual data.

The main purpose for the existence of DataSubset is to delay data access and movement until an actual batch of data (or single observation) is needed for some computation. This is particularily useful when the data is not located in memory, but on the hard drive or some remote location. In such a scenario one wants to load the required data only when needed.

This type is usually not constructed manually, but instead instantiated by calling datasubset, shuffleobs, or splitobs

In case data is some Tuple, the constructor will be mapped over its elements. That means that the constructor returns a Tuple of DataSubset instead of a DataSubset of Tuple.

Arguments

data : The object describing the dataset. Can be of any type as long as it implements getobs and nobs (see Details for more information).

indices : Optional. The index or indices of the observation(s) in data that the subset should represent. Can be of type Int or some subtype of AbstractVector.

obsdim : Optional. If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

Methods

getindex : Returns the observation(s) of the given index/indices as a new DataSubset. No data is copied aside from the required indices.

nobs : Returns the total number observations in the subset (not the whole data set underneath).

getobs : Returns the underlying data that the DataSubset represents at the given relative indices. Note that these indices are in "subset space", and in general will not directly correspond to the same indices in the underlying data set.

Details

For DataSubset to work on some data structure, the desired type MyType must implement the following interface:

LearnBase.getobs(data::MyType, idx, [obsdim::ObsDimension]) : Should return the observation(s) indexed by idx. In what form is up to the user. Note that idx can be of type Int or AbstractVector.

LearnBase.nobs(data::MyType, [obsdim::ObsDimension]) : Should return the total number of observations in data

The following methods can also be provided and are optional:

LearnBase.getobs(data::MyType) : By default this function is the identity function. If that is not the behaviour that you want for your type, you need to provide this method as well.

LearnBase.datasubset(data::MyType, idx, obsdim::ObsDimension) : If your custom type has its own kind of subset type, you can return it here. An example for such a case are SubArray for representing a subset of some AbstractArray. Note: If your type has no use for obsdim then dispatch on ::ObsDim.Undefined in the signature.

LearnBase.getobs!(buffer, data::MyType, [idx], [obsdim::ObsDimension]) : Inplace version of getobs(data, idx, obsdim). If this method is provided for MyType, then eachobs and eachbatch (among others) can preallocate a buffer that is then reused every iteration. Note: buffer should be equivalent to the return value of getobs(::MyType, ...), since this is how buffer is preallocated by default.

LearnBase.gettargets(data::MyType, idx, [obsdim::ObsDimension]) : If MyType has a special way to query targets without needing to invoke getobs, then you can provide your own logic here. This can be useful when the targets of your are always loaded as metadata, while the data itself remains on the hard disk until actually needed.

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Tom Breloff (Github: https://github.com/tbreloff)

Examples

X, y = MLDataUtils.load_iris()

# The iris set has 150 observations and 4 features
@assert size(X) == (4,150)

# Represents the 80 observations as a DataSubset
subset = DataSubset(X, 21:100)
@assert nobs(subset) == 80
@assert typeof(subset) <: DataSubset
# getobs indexes into the subset
@assert getobs(subset, 1:10) == X[:, 21:30]

# You can also work with data that uses some other dimension
# to denote the observations.
@assert size(X') == (150,4)
subset = DataSubset(X', 21:100, obsdim = :first) # or "obsdim = 1"
@assert nobs(subset) == 80

# To specify the obsdim in a type-stable way, use positional arguments
# provided by the submodule `ObsDim`.
@inferred DataSubset(X', 21:100, ObsDim.First())

# Subsets also works for tuple of data. (useful for labeled data)
subset = DataSubset((X,y), 1:100)
@assert nobs(subset) == 100
@assert typeof(subset) <: Tuple # Tuple of DataSubset

# The lowercase version tries to avoid boxing into DataSubset
# for types that provide a custom "subset", such as arrays.
# Here it instead creates a native SubArray.
subset = datasubset(X, 1:100)
@assert nobs(subset) == 100
@assert typeof(subset) <: SubArray

# Also works for tuples of arbitrary length
subset = datasubset((X,y), 1:100)
@assert nobs(subset) == 100
@assert typeof(subset) <: Tuple # tuple of SubArray

# Split dataset into training and test split
train, test = splitobs(shuffleobs((X,y)), at = 0.7)
@assert typeof(train) <: Tuple # of SubArray
@assert typeof(test)  <: Tuple # of SubArray
@assert nobs(train) == 105
@assert nobs(test) == 45

see also

datasubset, getobs, nobs, splitobs, shuffleobs, KFolds, BatchView, ObsView,

MLDataPattern.FoldsView — Type.

FoldsView(data, train_indices, val_indices, [obsdim])

Description

Create a vector-like representation of data where each individual element is partition of data in the form of a tuple of two data subsets (a training- and a validation subset).

The purpose of FoldsView is to apply a precomputed sequence of subset assignment indices to some data container in a convenient manner. By itself, FoldsView is agnostic to any particular repartitioning strategy (such as k-folds). Instead, the assignments, train_indices and val_indices, need to be precomputed by such a strategy and then passed to FoldsView with a concrete data container. The resulting object can then be queried for its individual folds using getindex, or simply iterated over.

Arguments

data : The object describing the dataset. Can be of any type as long as it implements getobs and nobs (see Details for more information).

train_indices : Vector of integer vectors containing the sequence of training assignments. This means that each element is a vector of indices that describe each training subset. The length of this vector must match val_indices.

val_indices : Vector of integer vectors containing the sequence of validation assignments. This means that each element is a vector of indices that describe each validation subset. The length of this vector must match train_indices.

obsdim : Optional. If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

Details

For FoldsView to work on some data structure, the desired type MyType must implement the following interface:

LearnBase.getobs(data::MyType, idx, [obsdim::ObsDimension]) : Should return the observation(s) indexed by idx. In what form is up to the user. Note that idx can be of type Int or AbstractVector.

LearnBase.nobs(data::MyType, [obsdim::ObsDimension]) : Should return the total number of observations in data

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Tom Breloff (Github: https://github.com/tbreloff)

Examples

# Load iris data for demonstration purposes
X, y = MLDataUtils.load_iris()

# Compute train- and validation-partition indices using kfolds
train_idx, val_idx = kfolds(nobs(X), 10)

# These two vectors contain the indices vector for each partitioning
@assert typeof(train_idx) <: Vector{Vector{Int64}}
@assert typeof(val_idx)   <: Vector{UnitRange{Int64}}
@assert length(train_idx) == length(val_idx) == 10

# Using the repartitioning as an oterator
for (train_X, val_X) in FoldsView(X, train_idx, val_idx)
    @assert size(train_X) == (4, 135)
    @assert size(val_X) == (4, 15)
end

# Calling kfolds with the dataset will create
# the FoldsView for you automatically.
# Thus this code is equivalent to above
for (train_X, val_X) in kfolds(X, 10)
    @assert size(train_X) == (4, 135)
    @assert size(val_X) == (4, 15)
end

# leavout is a shortcut for setting k = nobs(X)
for (train_X, val_X) in leaveout(X)
    @assert size(val_X) == (4, 1)
end

see also

kfolds, leaveout, splitobs, DataSubset

MLDataPattern.LabeledSlidingWindow

MLDataPattern.MLDataPattern

MLDataPattern.ObsDim

MLDataPattern.ObsView — Type.

ObsView(data, [obsdim])

Description

Create a view of the given data in the form of a vector of individual observations. Any data access is delayed until getindex is called, and even getindex returns the result of datasubset which in general avoids data movement until getobs is called.

If used as an iterator, the view will iterate over the dataset once, effectively denoting an epoch. Each iteration will return a lazy subset to the current observation.

Arguments

data : The object describing the dataset. Can be of any type as long as it implements getobs and nobs (see Details for more information).

obsdim : Optional. If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

Methods

Aside from the AbstractArray interface following additional methods are provided:

getobs(data::ObsView, indices::AbstractVector) : Returns a Vector of indivdual observations specified by indices.

nobs(data::ObsView) : Returns the number of observations in data that the iterator will go over.

Details

For ObsView to work on some data structure, the type of the given variable data must implement the data container interface. See ?DataSubset for more info.

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Tom Breloff (Github: https://github.com/tbreloff)

Examples

X, Y = MLDataUtils.load_iris()

A = obsview(X)
@assert typeof(A) <: ObsView <: AbstractVector
@assert eltype(A) <: SubArray{Float64,1}
@assert length(A) == 150 # Iris has 150 observations
@assert size(A[1]) == (4,) # Iris has 4 features

for x in obsview(X)
    @assert typeof(x) <: SubArray{Float64,1}
end

# iterate over each individual labeled observation
for (x,y) in obsview((X,Y))
    @assert typeof(x) <: SubArray{Float64,1}
    @assert typeof(y) <: String
end

# same but in random order
for (x,y) in obsview(shuffleobs((X,Y)))
    @assert typeof(x) <: SubArray{Float64,1}
    @assert typeof(y) <: String
end

see also

eachobs, BatchView, shuffleobs, getobs, nobs, DataSubset

MLDataPattern.RandomBatches — Type.

RandomBatches(data, [size], [count], [obsdim])

Description

Create an iterator that generates count randomly sampled batches from data with a batch-size of size . In the case count is not provided, it will generate random batches indefinitely.

Arguments

data : The object describing the dataset. Can be of any type as long as it implements getobs and nobs (see Details for more information).

size : Optional. The batch-size of each batch. I.e. the number of randomly sampled observations in each batch

count : Optional. The number of randomly sampled batches that the iterator will generate before stopping. If omitted, the iterator will generate randomly sampled batches forever.

obsdim : Optional. If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

Details

For RandomBatches to work on some data structure, the type of given variable data must implement the data container interface. See ?DataSubset for more info.

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Tom Breloff (Github: https://github.com/tbreloff)

Examples

X, Y = MLDataUtils.load_iris()

# go over 500 randomly sampled batches of batchsize 10
i = 0
for x in RandomBatches(X, 10, 500) # also: RandomObs(X, size = 10, count = 500)
    @assert typeof(x) <: SubArray{Float64,2}
    @assert size(x) == (4,10)
    i += 1
end
@assert i == 500

# if no count it provided the iterator will generate samples forever
for x in RandomBatches(X, 10)
    # this loop will never stop unless break is used
    if true; break; end
end

# also works for multiple data arguments (e.g. labeled data)
for (x,y) in RandomBatches((X,Y), 10, 500)
    @assert typeof(x) <: SubArray{Float64,2}
    @assert typeof(y) <: SubArray{String,1}
end

see also

RandomObs, BatchView, ObsView, shuffleobs, DataSubset, BufferGetObs

MLDataPattern.RandomObs — Type.

RandomObs(data, [count], [obsdim])

Description

Create an iterator that generates count randomly sampled observations from data. In the case count is not provided, it will generate random samples indefinitely.

Arguments

data : The object describing the dataset. Can be of any type as long as it implements getobs and nobs (see Details for more information).

count : Optional. The number of randomly sampled observations that the iterator will generate before stopping. If omitted, the iterator will generate randomly sampled observations forever.

obsdim : Optional. If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

Details

For RandomObs to work on some data structure, the type of the given variable data must implement the data container interface. See ?DataSubset for more info.

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Tom Breloff (Github: https://github.com/tbreloff)

Examples

X, Y = MLDataUtils.load_iris()

# go over 500 randomly sampled observations in X
i = 0
for x in RandomObs(X, 500) # also: RandomObs(X, count = 500)
    @assert typeof(x) <: SubArray{Float64,1}
    @assert length(x) == 4
    i += 1
end
@assert i == 500

# if no count it provided the iterator will generate samples forever
for x in RandomObs(X)
    # this loop will never stop unless break is used
    if true; break; end
end

# also works for multiple data arguments (e.g. labeled data)
for (x,y) in RandomObs((X,Y), count = 100)
    @assert typeof(x) <: SubArray{Float64,1}
    @assert typeof(y) <: String
end

see also

BalancedObs, RandomBatches, ObsView, BatchView, shuffleobs, DataSubset, BufferGetObs

MLDataPattern.SlidingWindow

MLDataPattern.UnlabeledSlidingWindow

MLDataPattern.WindowBatchView

MLDataPattern.__getobs

MLDataPattern._batchrange — Function.

Helper function to translate a batch-index into a range of observations.

MLDataPattern._check_nobs

MLDataPattern._check_nobs_error

MLDataPattern._check_windowargs

MLDataPattern._compute_batch_settings — Function.

Helper function to compute sensible and compatible values for the size and count

MLDataPattern._datastr

MLDataPattern._getobs

MLDataPattern._getobs_eltype

MLDataPattern._getobs_error

MLDataPattern._gettarget

MLDataPattern._gettargets

MLDataPattern._gettargets_dispatch_idx

MLDataPattern._getwindowbatch

MLDataPattern._ispos

MLDataPattern._length

MLDataPattern._length_str

MLDataPattern._next_idx

MLDataPattern._targets

MLDataPattern._toVal

MLDataPattern._view

MLDataPattern._windowsettings

MLDataPattern.allowcontainer

MLDataPattern.batchsize — Function.

batchsize(data) -> Int

Return the fixed size of each batch in data.

MLDataPattern.batchview

LearnBase.datasubset — Function.

datasubset(data, [indices], [obsdim])

Returns a lazy subset of the observations in data that correspond to the given indices. No data will be copied except of the indices. It is similar to calling DataSubset(data, [indices], [obsdim]), but returns a SubArray if the type of data is Array or SubArray. Furthermore, this function may be extended for custom types of data that also want to provide their own subset-type.

If instead you want to get the subset of observations corresponding to the given indices in their native type, use getobs.

If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

see DataSubset for more information.

MLDataPattern.default_batch_size

MLDataPattern.downsample

MLDataPattern.eachbatch — Function.

eachbatch(data, [size], [count], [obsdim])

Iterate over data one batch at a time. If supported by the type of data, a buffer will be preallocated and reused for memory efficiency.

IMPORTANT: Avoid using collect, because in general each iteration could return the same object with mutated values. If that behaviour is undesired use BatchView instead.

The (constant) batch-size can be either provided directly using size or indirectly using count, which derives size based on nobs. In the case that the size of the dataset is not dividable by the specified (or inferred) size, the remaining observations will be ignored.

X = rand(4,150)
for x in eachbatch(X, size = 10) # or: eachbatch(X, count = 15)
    # loop entered 15 times
    @assert typeof(x) <: Matrix{Float64}
    @assert size(x) == (4,10)
end

In the case of arrays it is assumed that the observations are represented by the last array dimension. This can be overwritten.

# This time flip the dimensions of the matrix
X = rand(150,4)
A = eachbatch(X, size = 10, obsdim = 1)
# The behaviour remains the same as before
@assert eltype(A) <: Array{Float64,2}
@assert length(A) == 15

Multiple variables are supported (e.g. for labeled data)

for (x,y) in eachbatch((X,Y))
    # ...
end

Note that internally eachbatch(data, ...) maps to BufferGetObs(batchview(data, ...)).

@assert typeof(eachbatch(X)) <: BufferGetObs
@assert typeof(eachbatch(X).iter) <: BatchView

This means that the following code:

for batch in eachbatch(data, batchsize, obsdim)
    # ...
end

is roughly equivalent to:

batch = getobs(data, collect(1:batchsize), obsdim) # use first element to preallocate buffer
for _ in batchview(data, batchsize, obsdim)
    getobs!(batch, _) # reuse buffer each iteration
    # ...
end

see BufferGetObs, batchview, and getobs! for more info. also see eachobs for a single-observation version.

MLDataPattern.eachobs — Function.

eachobs(data, [obsdim])

Iterate over data one observation at a time. If supported by the type of data, a buffer will be preallocated and reused for memory efficiency.

IMPORTANT: Avoid using collect, because in general each iteration could return the same object with mutated values. If that behaviour is undesired use obsview instead.

X = rand(4,100)
for x in eachobs(X)
    # loop entered 100 times
    @assert typeof(x) <: Vector{Float64}
    @assert size(x) == (4,)
end

In the case of arrays it is assumed that the observations are represented by the last array dimension. This can be overwritten.

# This time flip the dimensions of the matrix
X = rand(100,4)
A = eachobs(X, obsdim=1)
# The behaviour remains the same as before
@assert eltype(A) <: Array{Float64,1}
@assert length(A) == 100

Multiple variables are supported (e.g. for labeled data)

for (x,y) in eachobs((X,Y))
    # ...
end

Note that internally eachobs(data, obsdim) maps to BufferGetObs(obsview(data, obsdim)).

@assert typeof(eachobs(X)) <: BufferGetObs
@assert typeof(eachobs(X).iter) <: ObsView

This means that the following code:

for obs in eachobs(data, obsdim)
    # ...
end

is roughly equivalent to:

obs = getobs(data, 1, obsdim) # use first element to preallocate buffer
for _ in obsview(data, obsdim)
    getobs!(obs, _) # reuse buffer each iteration
    # ...
end

see BufferGetObs, obsview, and getobs! for more info. also see eachbatch for a mini-batch version.

MLDataPattern.eachtarget

MLDataPattern.eval

LearnBase.getobs — Function.

getobs(data, [idx], [obsdim])

Return the observation(s) in data that correspond to the given index/indices in idx. Note that idx can be of type Int or AbstractVector.

The returned observation(s) should be in the form intended to be passed as-is to some learning algorithm. There is no strict requirement that dictates what form or type that is. We do, however, expect it to be consistent for idx being an integer, as well as idx being an abstract vector, respectively.

If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

LearnBase.getobs! — Function.

getobs!(buffer, data, [idx], [obsdim]) -> buffer

Write the observation(s) from data that correspond to the given index/indices in idx into buffer. Note that idx can be of type Int or AbstractVector.

Unless explicitly implemented for data it defaults to returning getobs(data, idx, obsdim) in which case buffer is ignored.

If it makes sense for the type of data, obsdim can be used to specify which dimension of data denotes the observations. It can be specified in a type-stable manner as a positional argument (see ?LearnBase.ObsDim), or more conveniently as a smart keyword argument.

MLDataPattern.getobs_targetfun

MLDataPattern.include

MLDataPattern.kfolds — Function.

kfolds(n::Integer, [k = 5]) -> Tuple

Compute the train/validation assignments for k repartitions of n observations, and return them in the form of two vectors. The first vector contains the index-vectors for the training subsets, and the second vector the index-vectors for the validation subsets respectively. A general rule of thumb is to use either k = 5 or k = 10. The following code snippet generates the indices assignments for k = 5

julia> train_idx, val_idx = kfolds(10, 5);

Each observation is assigned to the validation subset once (and only once). Thus, a union over all validation index-vectors reproduces the full range 1:n. Note that there is no random assignment of observations to subsets, which means that adjacent observations are likely to be part of the same validation subset.

julia> train_idx
5-element Array{Array{Int64,1},1}:
 [3,4,5,6,7,8,9,10]
 [1,2,5,6,7,8,9,10]
 [1,2,3,4,7,8,9,10]
 [1,2,3,4,5,6,9,10]
 [1,2,3,4,5,6,7,8]

julia> val_idx
5-element Array{UnitRange{Int64},1}:
 1:2
 3:4
 5:6
 7:8
 9:10

kfolds(data, [k = 5], [obsdim]) -> FoldsView

Repartition a data container k times using a k folds strategy and return the sequence of folds as a lazy FoldsView. The resulting FoldsView can then be indexed into or iterated over. Either way, only data subsets are created. That means that no actual data is copied until getobs is invoked.

Conceptually, a k-folds repartitioning strategy divides the given data into k roughly equal-sized parts. Each part will serve as validation set once, while the remaining parts are used for training. This results in k different partitions of data.

In the case that the size of the dataset is not dividable by the specified k, the remaining observations will be evenly distributed among the parts.

for (x_train, x_val) in kfolds(X, k = 10)
    # code called 10 times
    # nobs(x_val) may differ up to ±1 over iterations
end

Multiple variables are supported (e.g. for labeled data)

for ((x_train, y_train), val) in kfolds((X, Y), k = 10)
    # ...
end

By default the folds are created using static splits. Use shuffleobs to randomly assign observations to the folds.

for (x_train, x_val) in kfolds(shuffleobs(X), k = 10)
    # ...
end

see FoldsView for more info, or leaveout for a related function.

MLDataPattern.leaveout — Function.

leaveout(n::Integer, [size = 1]) -> Tuple

Compute the train/validation assignments for k ≈ n/size repartitions of n observations, and return them in the form of two vectors. The first vector contains the index-vectors for the training subsets, and the second vector the index-vectors for the validation subsets respectively. Each validation subset will have either size or size+1 observations assigned to it. The following code snippet generates the index-vectors for size = 2.

julia> train_idx, val_idx = leaveout(10, 2);

Each observation is assigned to the validation subset once (and only once). Thus, a union over all validation index-vectors reproduces the full range 1:n. Note that there is no random assignment of observations to subsets, which means that adjacent observations are likely to be part of the same validation subset.

julia> train_idx
5-element Array{Array{Int64,1},1}:
 [3,4,5,6,7,8,9,10]
 [1,2,5,6,7,8,9,10]
 [1,2,3,4,7,8,9,10]
 [1,2,3,4,5,6,9,10]
 [1,2,3,4,5,6,7,8]

julia> val_idx
5-element Array{UnitRange{Int64},1}:
 1:2
 3:4
 5:6
 7:8
 9:10

leaveout(data, [size = 1], [obsdim]) -> FoldsView

Repartition a data container using a k-fold strategy, where k is chosen in such a way, that each validation subset of the resulting folds contains roughly size observations. Defaults to size = 1, which is also known as "leave-one-out" partitioning.

The resulting sequence of folds is returned as a lazy FoldsView, which can be index into or iterated over. Either way, only data subsets are created. That means no actual data is copied until getobs is invoked.

for (train, val) in leaveout(X, size = 2)
    # if nobs(X) is dividable by 2,
    # then nobs(val) will be 2 for each iteraton,
    # otherwise it may be 3 for the first few iterations.
end

see FoldsView for more info, or kfolds for a related function.

StatsBase.nobs — Function.

nobs(data, [obsdim]) -> Int

Return the total number of observations contained in data.

The optional parameter obsdim can be used to specify which dimension denotes the observations, if that concept makes sense for the type of data. See ?LearnBase.ObsDim for more information.

MLDataPattern.obsdim_string

MLDataPattern.obsview

MLDataPattern.oversample — Function.

oversample([f], data, [fraction = 1], [shuffle = true], [obsdim])

Generate a re-balanced version of data by repeatedly sampling existing observations in such a way that every class will have at least fraction times the number observations of the largest class. This way, all classes will have a minimum number of observations in the resulting data set relative to what largest class has in the given (original) data.

As an example, by default (i.e. with fraction = 1) the resulting dataset will be near perfectly balanced. On the other hand, with fraction = 0.5 every class in the resulting data with have at least 50% as many observations as the largest class.

The convenience parameter shuffle determines if the resulting data will be shuffled after its creation; if it is not shuffled then all the repeated samples will be together at the end, sorted by class. Defaults to true.

The optional parameter obsdim can be used to specify which dimension denotes the observations, if that concept makes sense for the type of data. See ?ObsDim for more information.

# 6 observations with 3 features each
X = rand(3, 6)
# 2 classes, severely imbalanced
Y = ["a", "b", "b", "b", "b", "a"]

# oversample the class "a" to match "b"
X_bal, Y_bal = oversample((X,Y))

# this results in a bigger dataset with repeated data
@assert size(X_bal) == (3,8)
@assert length(Y_bal) == 8

# now both "a", and "b" have 4 observations each
@assert sum(Y_bal .== "a") == 4
@assert sum(Y_bal .== "b") == 4

For this function to work, the type of data must implement nobs and getobs. For example, the following code allows oversample to work on a DataTable.

# Make DataTables.jl work
LearnBase.getobs(data::DataTable, i) = data[i,:]
LearnBase.nobs(data::DataTable) = nrow(data)

You can use the parameter f to specify how to extract or retrieve the targets from each observation of the given data. Note that if data is a tuple, then it will be assumed that the last element of the tuple contains the targets and f will be applied to each observation in that element.

julia> data = DataTable(Any[rand(6), rand(6), [:a,:b,:b,:b,:b,:a]], [:X1,:X2,:Y])
6×3 DataTables.DataTable
│ Row │ X1        │ X2          │ Y │
├─────┼───────────┼─────────────┼───┤
│ 1   │ 0.226582  │ 0.0443222   │ a │
│ 2   │ 0.504629  │ 0.722906    │ b │
│ 3   │ 0.933372  │ 0.812814    │ b │
│ 4   │ 0.522172  │ 0.245457    │ b │
│ 5   │ 0.505208  │ 0.11202     │ b │
│ 6   │ 0.0997825 │ 0.000341996 │ a │

julia> getobs(oversample(row->row[:Y], data))
8×3 DataTables.DataTable
│ Row │ X1        │ X2          │ Y │
├─────┼───────────┼─────────────┼───┤
│ 1   │ 0.0997825 │ 0.000341996 │ a │
│ 2   │ 0.505208  │ 0.11202     │ b │
│ 3   │ 0.226582  │ 0.0443222   │ a │
│ 4   │ 0.0997825 │ 0.000341996 │ a │
│ 5   │ 0.504629  │ 0.722906    │ b │
│ 6   │ 0.522172  │ 0.245457    │ b │
│ 7   │ 0.226582  │ 0.0443222   │ a │
│ 8   │ 0.933372  │ 0.812814    │ b │

see DataSubset for more information on data subsets.

see also undersample and stratifiedobs.

MLDataPattern.randobs — Function.

randobs(data, [n], [obsdim])

Pick a random observation or a batch of n random observations from data.

The optional (keyword) parameter obsdim allows one to specify which dimension denotes the observations. see LearnBase.ObsDim for more detail.

For this function to work, the type of data must implement nobs and getobs.

MLDataPattern.shuffleobs — Function.

shuffleobs(data, [obsdim])

Return a "subset" of data that spans all observations, but has the order of the observations shuffled.

The values of data itself are not copied. Instead only the indices are shuffled. This function calls datasubset to accomplish that, which means that the return value is likely of a different type than data.

# For Arrays the subset will be of type SubArray
@assert typeof(shuffleobs(rand(4,10))) <: SubArray

# Iterate through all observations in random order
for (x) in eachobs(shuffleobs(X))
    ...
end

The optional (keyword) parameter obsdim allows one to specify which dimension denotes the observations. see LearnBase.ObsDim for more detail.

For this function to work, the type of data must implement nobs and getobs. See DataSubset for more information.

MLDataPattern.slidingwindow — Function.

slidingwindow(data, size, [stride], [obsdim]) -> UnlabeledSlidingWindow

Return a vector-like view of the data for which each element is a fixed size "window" of size adjacent observations. By default these windows are not overlapping. Note that only complete windows are included in the output, which implies that it is possible for excess observations to be omitted from the view.

julia> A = slidingwindow(1:20, 6)
3-element slidingwindow(::UnitRange{Int64}, 6) with element type SubArray{Int64,1,UnitRange{Int64},Tuple{UnitRange{Int64}},true}:
 [1, 2, 3, 4, 5, 6]
 [7, 8, 9, 10, 11, 12]
 [13, 14, 15, 16, 17, 18]

Note that the values of data itself are not copied. Instead the function datasubset is called when getindex is invoked. To actually get a copy of the data at some window use the function getobs.

julia> A[1]
6-element SubArray{Int64,1,UnitRange{Int64},Tuple{UnitRange{Int64}},true}:
 1
 ⋮
 6

julia> getobs(A, 1)
6-element Array{Int64,1}:
 1
 ⋮
 6

The optional parameter stride can be used to specify the distance between the start elements of each adjacent window. By default the stride is equal to the window size.

julia> slidingwindow(1:20, 6, stride = 3)
5-element slidingwindow(::UnitRange{Int64}, 6, stride = 3) with element type SubArray{Int64,1,UnitRange{Int64},Tuple{UnitRange{Int64}},true}:
 [1, 2, 3, 4, 5, 6]
 [4, 5, 6, 7, 8, 9]
 [7, 8, 9, 10, 11, 12]
 [10, 11, 12, 13, 14, 15]
 [13, 14, 15, 16, 17, 18]

The optional (keyword) parameter obsdim allows one to specify which dimension denotes the observations. see LearnBase.ObsDim for more detail.

slidingwindow(f, data, size, [stride], [excludetarget], [obsdim]) -> LabeledSlidingWindow

Return a vector-like view of the data for which each element is a tuple of two elements:

A fixed size "window" of size adjacent observations. By default these windows are not overlapping. This can be changed by explicitly specifying a stride.

A single target (or vector of targets) for the window. The content of the target(s) is defined by the label-index function f.

The label-index function f is a unary function that takes the index of the first observation in the current window and should return the index (or indices) of the associated target(s) for that window.

julia> A = slidingwindow(i->i+6, 1:20, 6)
3-element slidingwindow(::##3#4, ::UnitRange{Int64}, 6) with element type Tuple{...}
 ([1, 2, 3, 4, 5, 6], 7)
 ([7, 8, 9, 10, 11, 12], 13)
 ([13, 14, 15, 16, 17, 18], 19)

Note that only complete and in-bound windows are included in the output, which implies that it is possible for excess observations to be omitted from the resulting view.

julia> A = slidingwindow(i->i-1, 1:20, 6)
2-element slidingwindow(::##5#6, ::UnitRange{Int64}, 6) with element type Tuple{...}
 ([7, 8, 9, 10, 11, 12], 6)
 ([13, 14, 15, 16, 17, 18], 12)

As hinted above, it is also allowed for f to return a vector of indices. This can be useful for emulating techniques such as skip-gram.

julia> data = split("The quick brown fox jumps over the lazy dog")
9-element Array{SubString{String},1}:
 "The"
 "quick"
 ⋮
 "lazy"
 "dog"

julia> A = slidingwindow(i->[i-2:i-1; i+1:i+2], data, 1)
5-element slidingwindow(::##11#12, ::Array{SubString{String},1}, 1) with element type Tuple{...}:
 (["brown"], ["The", "quick", "fox", "jumps"])
 (["fox"], ["quick", "brown", "jumps", "over"])
 (["jumps"], ["brown", "fox", "over", "the"])
 (["over"], ["fox", "jumps", "the", "lazy"])
 (["the"], ["jumps", "over", "lazy", "dog"])

Should it so happen that the targets overlap with the features, then the affected observation(s) will be present in both. To change this behaviour one can set the optional parameter excludetarget = true. This will remove the target(s) from the feature window.

julia> slidingwindow(i->i+2, data, 5, stride = 1, excludetarget = true)
5-element slidingwindow(::##17#18, ::Array{SubString{String},1}, 5, stride = 1) with element type Tuple{...}:
 (["The", "quick", "fox", "jumps"], "brown")
 (["quick", "brown", "jumps", "over"], "fox")
 (["brown", "fox", "over", "the"], "jumps")
 (["fox", "jumps", "the", "lazy"], "over")
 (["jumps", "over", "lazy", "dog"], "the")

The optional (keyword) parameter obsdim allows one to specify which dimension denotes the observations. see LearnBase.ObsDim for more detail.

MLDataPattern.splitobs — Function.

splitobs(n::Int, [at = 0.7]) -> Tuple

Pre-compute the indices for two disjoint subsets and return them as a tuple of two ranges. The first range will span the first at fraction of possible indices, while the second range will cover the rest. These indices are applicable to any data container of size n.

julia> splitobs(100, at = 0.7)
# (1:70,71:100)

splitobs(data, [at = 0.7], [obsdim]) -> Tuple

Split the data into multiple subsets proportional to the value(s) of at.

Note that this function will perform the splits statically and thus not perform any randomization. The function creates a NTuple of data subsets in which the first N-1 elements/subsets contain the fraction of observations of data that is specified by at.

For example, if at is a Float64 then the return-value will be a tuple with two elements (i.e. subsets), in which the first element contains the fracion of observations specified by at and the second element contains the rest. In the following code the first subset train will contain the first 70% of the observations and the second subset test the rest.

train, test = splitobs(X, at = 0.7)

If at is a tuple of Float64 then additional subsets will be created. In this example train will have the first 50% of the observations, val will have next 30%, and test the last 20%

train, val, test = splitobs(X, at = (0.5, 0.3))

It is also possible to call splitobs with multiple data arguments as tuple, which all must have the same number of total observations. This is useful for labeled data.

train, test = splitobs((X, y), at = 0.7)
(x_train,y_train), (x_test,y_test) = splitobs((X, y), at = 0.7)

If the observations should be randomly assigned to a subset, then you can combine the function with shuffleobs

# This time observations are randomly assigned.
train, test = splitobs(shuffleobs((X,y)), at = 0.7)

When working with arrays one may want to choose which dimension represents the observations. By default the last dimension is assumed, but this can be overwritten.

# Here we say each row represents an observation
train, test = splitobs(X, obsdim = 1)

The functions also provide a type-stable API

# By avoiding keyword arguments, the compiler can infer the return type
train, test = splitobs((X,y), 0.7)
train, test = splitobs((X,y), 0.7, ObsDim.First())

see DataSubset for more information on data subsets.

see stratifiedobs for a related function that preserves the target distribution.

MLDataPattern.stratifiedobs — Function.

stratifiedobs([f], data, [p = 0.7], [shuffle = true], [obsdim]) -> Tuple

Partition the data into multiple disjoint subsets proportional to the value(s) of p. The observations are assignmed to a data subset using stratified sampling without replacement. These subsets are then returned as a Tuple of subsets, where the first element contains the fraction of observations of data that is specified by the first float in p.

For example, if p is a Float64 itself, then the return-value will be a tuple with two elements (i.e. subsets), in which the first element contains the fraction of observations specified by p and the second element contains the rest. In the following code the first subset train will contain around 70% of the observations and the second subset test the rest. The key difference to splitobs is that the class distribution in y will actively be preserved in train and test.

train, test = stratifiedobs(y, p = 0.7)

If p is a tuple of Float64 then additional subsets will be created. In this example train will contain about 50% of the observations, val will contain around 30%, and test the remaining 20%.

train, val, test = stratifiedobs(y, p = (0.5, 0.3))

It is also possible to call stratifiedobs with multiple data arguments as tuple, which all must have the same number of total observations. Note that if data is a tuple, then it will be assumed that the last element of the tuple contains the targets.

train, test = stratifiedobs((X, y), p = 0.7)
(X_train,y_train), (X_test,y_test) = stratifiedobs((X, y), p = 0.7)

The optional parameter shuffle determines if the resulting data subsets should be shuffled. If false, then the observations in the subsets will be grouped together according to their labels.

julia> y = ["a", "b", "b", "b", "b", "a"] # 2 imbalanced classes
6-element Array{String,1}:
 "a"
 "b"
 "b"
 "b"
 "b"
 "a"

julia> train, test = stratifiedobs(y, p = 0.5, shuffle = false)
(String["b","b","a"],String["b","b","a"])

The optional parameter obsdim can be used to specify which dimension denotes the observations, if that concept makes sense for the type of data. See ?ObsDim for more information.

# 2 imbalanced classes in one-of-k encoding
julia> X = [1 0; 1 0; 1 0; 1 0; 0 1; 0 1]
6×2 Array{Int64,2}:
 1  0
 1  0
 1  0
 1  0
 0  1
 0  1

julia> train, test = stratifiedobs(argmax, X, p = 0.5, obsdim = 1)
([1 0; 1 0; 0 1], [0 1; 1 0; 1 0])

For this function to work, the type of data must implement nobs and getobs. For example, the following code allows stratifiedobs to work on a DataTable.

# Make DataTables.jl work
LearnBase.getobs(data::DataTable, i) = data[i,:]
LearnBase.nobs(data::DataTable) = nrow(data)

You can use the parameter f to specify how to extract or retrieve the targets from each observation of the given data.

julia> data = DataTable(Any[rand(6), rand(6), [:a,:b,:b,:b,:b,:a]], [:X1,:X2,:Y])
6×3 DataTables.DataTable
│ Row │ X1        │ X2          │ Y │
├─────┼───────────┼─────────────┼───┤
│ 1   │ 0.226582  │ 0.0443222   │ a │
│ 2   │ 0.504629  │ 0.722906    │ b │
│ 3   │ 0.933372  │ 0.812814    │ b │
│ 4   │ 0.522172  │ 0.245457    │ b │
│ 5   │ 0.505208  │ 0.11202     │ b │
│ 6   │ 0.0997825 │ 0.000341996 │ a │

julia> train, test = stratifiedobs(row->row[:Y], data, 0.5);

julia> getobs(train)
3×3 DataTables.DataTable
│ Row │ X1        │ X2          │ Y │
├─────┼───────────┼─────────────┼───┤
│ 1   │ 0.933372  │ 0.812814    │ b │
│ 2   │ 0.522172  │ 0.245457    │ b │
│ 3   │ 0.0997825 │ 0.000341996 │ a │

julia> getobs(test)
3×3 DataTables.DataTable
│ Row │ X1       │ X2        │ Y │
├─────┼──────────┼───────────┼───┤
│ 1   │ 0.504629 │ 0.722906  │ b │
│ 2   │ 0.226582 │ 0.0443222 │ a │
│ 3   │ 0.505208 │ 0.11202   │ b │

see DataSubset for more information on data subsets.

see also undersample, oversample, splitobs.

LearnBase.targets — Function.

targets([f], data, [obsdim])

Extract the concrete targets from data and return them.

This function is eager in the sense that it will always call getobs unless a custom method for gettargets is implemented for the type of data. This will make sure that actual values are returned (in contrast to placeholders such as DataSubset or SubArray).

julia> targets(DataSubset([1,2,3]))
3-element Array{Int64,1}:
 1
 2
 3

If data is a tuple, then the convention is that the last element of the tuple contains the targets and the function is recursed once (and only once).

julia> targets(([1,2], [3,4]))
2-element Array{Int64,1}:
 3
 4

julia> targets(([1,2], ([3,4], [5,6])))
([3,4],[5,6])

If f is provided, then gettarget will be applied to each observation in data and the results will be returned as a vector.

julia> targets(argmax, [1 0 1; 0 1 0])
3-element Array{Int64,1}:
 1
 2
 1

The optional parameter obsdim can be used to specify which dimension denotes the observations, if that concept makes sense for the type of data. See ?ObsDim for more information.

julia> targets(argmax, [1 0; 0 1; 1 0], obsdim=1)
3-element Array{Int64,1}:
 1
 2
 1

MLDataPattern.undersample — Function.

undersample([f], data, [shuffle = false], [obsdim])

Generate a class-balanced version of data by subsampling its observations in such a way that the resulting number of observations will be the same number for every class. This way, all classes will have as many observations in the resulting data set as the smallest class has in the given (original) data.

The convenience parameter shuffle determines if the resulting data will be shuffled after its creation; if it is not shuffled then all the observations will be in their original order. Defaults to false.

The optional parameter obsdim can be used to specify which dimension denotes the observations, if that concept makes sense for the type of data. See ?ObsDim for more information.

# 6 observations with 3 features each
X = rand(3, 6)
# 2 classes, severely imbalanced
Y = ["a", "b", "b", "b", "b", "a"]

# subsample the class "b" to match "a"
X_bal, Y_bal = undersample((X,Y))

# this results in a smaller dataset
@assert size(X_bal) == (3,4)
@assert length(Y_bal) == 4

# now both "a", and "b" have 2 observations each
@assert sum(Y_bal .== "a") == 2
@assert sum(Y_bal .== "b") == 2

For this function to work, the type of data must implement nobs and getobs. For example, the following code allows undersample to work on a DataTable.

# Make DataTables.jl work
LearnBase.getobs(data::DataTable, i) = data[i,:]
LearnBase.nobs(data::DataTable) = nrow(data)

You can use the parameter f to specify how to extract or retrieve the targets from each observation of the given data. Note that if data is a tuple, then it will be assumed that the last element of the tuple contains the targets and f will be applied to each observation in that element.

julia> data = DataTable(Any[rand(6), rand(6), [:a,:b,:b,:b,:b,:a]], [:X1,:X2,:Y])
6×3 DataTables.DataTable
│ Row │ X1        │ X2          │ Y │
├─────┼───────────┼─────────────┼───┤
│ 1   │ 0.226582  │ 0.0443222   │ a │
│ 2   │ 0.504629  │ 0.722906    │ b │
│ 3   │ 0.933372  │ 0.812814    │ b │
│ 4   │ 0.522172  │ 0.245457    │ b │
│ 5   │ 0.505208  │ 0.11202     │ b │
│ 6   │ 0.0997825 │ 0.000341996 │ a │

julia> getobs(undersample(row->row[:Y], data))
4×3 DataTables.DataTable
│ Row │ X1        │ X2          │ Y │
├─────┼───────────┼─────────────┼───┤
│ 1   │ 0.226582  │ 0.0443222   │ a │
│ 2   │ 0.504629  │ 0.722906    │ b │
│ 3   │ 0.522172  │ 0.245457    │ b │
│ 4   │ 0.0997825 │ 0.000341996 │ a │

see DataSubset for more information on data subsets.

see also oversample and stratifiedobs.

MLDataPattern.upsample