Docstrings

SemidefiniteOptInterface.AF

SemidefiniteOptInterface.ASF

SemidefiniteOptInterface.AVF

SemidefiniteOptInterface.AbstractBlockMatrix

SemidefiniteOptInterface.AbstractSDOptimizer

SemidefiniteOptInterface.BlockMatrix

SemidefiniteOptInterface.CI

SemidefiniteOptInterface.DS

SemidefiniteOptInterface.MOI

SemidefiniteOptInterface.MOIU

SemidefiniteOptInterface.MockSDOptimizer

SemidefiniteOptInterface.NS

SemidefiniteOptInterface.PS

SemidefiniteOptInterface.SAF

SemidefiniteOptInterface.SDOIOptimizer

SemidefiniteOptInterface.SOItoMOIBridge

SemidefiniteOptInterface.SVF

SemidefiniteOptInterface.SemidefiniteOptInterface

SemidefiniteOptInterface.SolverStatus

SemidefiniteOptInterface.SupportedSets

SemidefiniteOptInterface.VAF

SemidefiniteOptInterface.VF

SemidefiniteOptInterface.VI

SemidefiniteOptInterface.VIS

SemidefiniteOptInterface.VVF

SemidefiniteOptInterface.ZS

SemidefiniteOptInterface._allocate_constraint

SemidefiniteOptInterface._constraintvariable!

SemidefiniteOptInterface._enumerate

SemidefiniteOptInterface._getattribute

SemidefiniteOptInterface._getblock

SemidefiniteOptInterface._getconstant

SemidefiniteOptInterface._length

SemidefiniteOptInterface._var

SemidefiniteOptInterface.addblkconstant

SemidefiniteOptInterface.addsetconstant

SemidefiniteOptInterface.block

SemidefiniteOptInterface.coefficienttype — Function.

coefficienttype(optimizer::AbstractSDOptimizer)

Returns the coefficienttype that should be used for optimizer.

SemidefiniteOptInterface.createslack!

SemidefiniteOptInterface.eval

SemidefiniteOptInterface.getX — Function.

getX(optimizer::AbstractSDOptimizer)

Returns the solution X as a block matrix.

SemidefiniteOptInterface.getZ — Function.

getZ(optimizer::AbstractSDOptimizer)

Returns the solution Z.

SemidefiniteOptInterface.getblock

SemidefiniteOptInterface.getblockdimension — Function.

getblockdimension(optimizer::AbstractSDOptimizer, blk)

Returns the dimension of the block blk.

SemidefiniteOptInterface.getconstraintcoefficients — Function.

getconstraintcoefficients(optimizer::AbstractSDOptimizer, constr::Integer)

Return the list of entries blk, i, j of the matrix of the constraint constr.

SemidefiniteOptInterface.getconstraintconstant — Function.

getconstraintconstant(optimizer::AbstractSDOptimizer, constr::Integer)

Sets the entry constr of b to val.

SemidefiniteOptInterface.getdual

SemidefiniteOptInterface.getdualobjectivevalue — Function.

getdualobjectivevalue(optimizer::AbstractSDOptimizer)

Returns the dual objective value.

SemidefiniteOptInterface.getmatdim

SemidefiniteOptInterface.getnumberofblocks — Function.

getnumberofblocks(optimizer::AbstractSDOptimizer)

Returns the number of blocks of the block matrix.

SemidefiniteOptInterface.getnumberofconstraints — Function.

getnumberofconstraints(optimizer::AbstractSDOptimizer)

Returns the number of constraints of the model.

SemidefiniteOptInterface.getobjectivecoefficients — Function.

getobjectivecoefficients(optimizer::AbstractSDOptimizer)

Return the list of entries blk, i, j of the objective matrix.

SemidefiniteOptInterface.getprimalobjectivevalue — Function.

getprimalobjectivevalue(optimizer::AbstractSDOptimizer)

Returns the primal objective value.

SemidefiniteOptInterface.getslack

SemidefiniteOptInterface.getvardual

SemidefiniteOptInterface.getvarprimal

SemidefiniteOptInterface.gety — Function.

gety(optimizer::AbstractSDOptimizer)

Returns the solution y.

SemidefiniteOptInterface.include

SemidefiniteOptInterface.init! — Function.

init!(optimizer::AbstractSDOptimizer, blkdims::Vector{Int}, nconstrs::Integer)

Initialize the optimizer with nconstrs constraints and blkdims blocks.

SemidefiniteOptInterface.isfree

SemidefiniteOptInterface.load_objective_term!

SemidefiniteOptInterface.loadcoefficients!

SemidefiniteOptInterface.loadfreevariables!

SemidefiniteOptInterface.loadslack!

SemidefiniteOptInterface.loadslacks!

SemidefiniteOptInterface.mockSDoptimizer

SemidefiniteOptInterface.nblocks

SemidefiniteOptInterface.nconstraints

SemidefiniteOptInterface.newblock

SemidefiniteOptInterface.nextline

SemidefiniteOptInterface.output_index

SemidefiniteOptInterface.readsdpa!

SemidefiniteOptInterface.scalar_term

SemidefiniteOptInterface.scalevec!

SemidefiniteOptInterface.setconstant!

SemidefiniteOptInterface.setconstraintcoefficient! — Function.

setconstraintcoefficient!(optimizer::AbstractSDOptimizer, val, constr::Integer, blk::Integer, i::Integer, j::Integer)

Set the entry i, j of the block blk of the matrix of the constraint constr to val.

SemidefiniteOptInterface.setconstraintconstant! — Function.

setconstraintconstant!(optimizer::AbstractSDOptimizer, val, constr::Integer)

Get the entry constr of b.

SemidefiniteOptInterface.setobjectivecoefficient! — Function.

setobjectivecoefficient!(optimizer::AbstractSDOptimizer, val, blk::Integer, i::Integer, j::Integer)

Set the entry i, j of the block blk of the objective matrix to val.

SemidefiniteOptInterface.setvarmap!

SemidefiniteOptInterface.sympackedlen

SemidefiniteOptInterface.to_matrix

SemidefiniteOptInterface.unfree

SemidefiniteOptInterface.varmap

SemidefiniteOptInterface.vscaling

SemidefiniteOptInterface.writesdpa