Docstrings

Humanize.BASES

Humanize.Humanize

Humanize.SUFFIXES

Humanize.datasize — Function.

datasize(value::Number; style=:dec, format="%.1f")

Format a number of bytes in a human-friendly format (eg. 10 kB).

style=:dec - default, decimal suffixes (kB, MB), base 10^3
style=:bin - binary suffixes (KiB, MiB), base 2^10
style=:gnu - GNU-style (ls -sh style, K, M), base 2^10

Humanize.digitsep — Function.

digitsep(value::Integer, separator=",", per_separator=3)

Convert an integer to a string, separating each per_separator digits by separator.

digitsep(value)  # 12345678 -> "12,345,678".
digitsep(value, separator="'")  # 12345678 -> "12'345'678".
digitsep(value, separator="/", per_separator=4)  # 12345678 -> "1234/5678".

Humanize.eval

Humanize.include

Humanize.timedelta — Function.

timedelta(secs::Integer)
timedelta(seconds::Dates.Second)
timedelta(Δdt::Dates.Millisecond)
timedelta(Δdate::Dates.Day)

Format a time length in a human friendly format.

timedelta(seconds::Integer)  # 3699 -> "An hour".
timedelta(Δdt::Dates.Millisecond)
    e.g. DateTime(2014,2,3) - DateTime(2013,3,7) -> "11 months".
timedelta(Δdate::Dates.Day)
    e.g. Date(2014,3,7) - Date(2013,2,4) -> "1 year, 1 month".