Docstrings

ScikitLearn.@reexportsk — Macro.

@reexportsk(identifiers...)

is equivalent to using Skcore: identifiers... export identifiers...

ScikitLearn.Skcore.@sk_import — Macro.

@skimport imports models from the Python version of scikit-learn. For instance, the Julia equivalent of `from sklearn.linearmodel import LinearRegression, LogicisticRegression` is:

@sk_import linear_model: (LinearRegression, LogisticRegression)
model = fit!(LinearRegression(), X, y)

ScikitLearn.CrossValidation

ScikitLearn.DataFrameColSelector — Type.

DataFrameColSelector(colums::Vector{Symbol}; output_type=Matrix{Float64}

This is a pared-down, less featureful version of DataFrameMapper, but it also runs faster. It only allows selecting columns from the input DataFrame. Use in a pipeline.

ScikitLearn.DataFrameMapper — Type.

DataFrameMapper(features; sparse=false, missing2NaN=false,
                output_type=Matrix{Float64})

Map DataFrame column subsets to their own scikit-learn transformation.

Arguments:

features: a vector of pairs. The first element is the column name. This can be a symbol (for one column) or a vector of symbols. The second element is an object that supports sklearn's transform interface, or a vector of such objects.

sparse will return sparse matrix if set to true and any of the extracted features is sparse. Defaults to False.

missing2NaN will convert missing values to NaNs (necessary for Python models)

output_type: the type of the result (default: Matrix{Float64})

ScikitLearn.GridSearch

ScikitLearn.Models

ScikitLearn.Pipelines

ScikitLearn.Preprocessing

ScikitLearn.ScikitLearn

ScikitLearn.Skcore

ScikitLearn.Utils

ScikitLearn._build_transformer

ScikitLearn._get_col_subset — Function.

Get a subset of columns from the given table X. X a dataframe; the table to select columns from cols a symbol or vector of symbols representing the columns to select Returns a matrix with the data from the selected columns

ScikitLearn._handle_feature — Function.

Convert a vector to a matrix

ScikitLearn._maybe_convert_missing

ScikitLearn._transform

ScikitLearn.clone

ScikitLearnBase.decision_function — Function.

Applies transforms to the data, and the decisionfunction method of the final estimator. Valid only if the final estimator implements decisionfunction.

Parameters

X : iterable Data to predict on. Must fulfill input requirements of first step of the pipeline.

ScikitLearn.dummy_input

ScikitLearn.eval

ScikitLearnBase.fit! — Function.

Fit all transformers using X.

Parameters

X : array-like or sparse matrix, shape (nsamples, nfeatures) Input data, used to fit transformers.

Run fit with all sets of parameters.

Parameters

X : array-like, shape = [nsamples, nfeatures] Training vector, where nsamples is the number of samples and nfeatures is the number of features.

y : array-like, shape = [nsamples] or [nsamples, n_output], optional Target relative to X for classification or regression; None for unsupervised learning.

Run fit on the estimator with randomly drawn parameters.

Parameters

X : array-like, shape = [nsamples, nfeatures] Training vector, where nsamples in the number of samples and nfeatures is the number of features.

y : array-like, shape = [nsamples] or [nsamples, n_output], optional Target relative to X for classification or regression; None for unsupervised learning.

ScikitLearn.fit_predict!

ScikitLearn.fit_transform!

ScikitLearn.get_classes

ScikitLearn.get_components

ScikitLearnBase.get_feature_names — Function.

Get feature names from all transformers.

Returns

feature_names : list of strings Names of the features produced by transform.

ScikitLearn.get_params

ScikitLearn.include

ScikitLearnBase.inverse_transform — Function.

Applies inverse transform to the data. Starts with the last step of the pipeline and applies $inverse_transform$ in inverse order of the pipeline steps. Valid only if all steps of the pipeline implement inverse_transform.

Parameters

X : iterable Data to inverse transform. Must fulfill output requirements of the last step of the pipeline.

ScikitLearn.is_classifier

ScikitLearn.is_pairwise

ScikitLearn.partial_fit!

ScikitLearn.predict

ScikitLearn.predict_dist

ScikitLearn.predict_log_proba

ScikitLearn.predict_proba

ScikitLearn.sample

ScikitLearnBase.score — Function.

Applies transforms to the data, and the score method of the final estimator. Valid only if the final estimator implements score.

Parameters

X : iterable Data to score. Must fulfill input requirements of first step of the pipeline.

y : iterable, default=None Targets used for scoring. Must fulfill label requirements for all steps of the pipeline.

ScikitLearn.score_samples

ScikitLearn.set_params!

ScikitLearnBase.transform — Function.

Transform X separately by each transformer, concatenate results.

Parameters

X : array-like or sparse matrix, shape (nsamples, nfeatures) Input data to be transformed.

Returns

Xt : array-like or sparse matrix, shape (nsamples, sumncomponents) hstack of results of transformers. sumncomponents is the sum of n_components (output dimension) over transformers.