Docstrings

ImplicitEquations.BInterval

ImplicitEquations.Eq

ImplicitEquations.FALSE

ImplicitEquations.GRAPH — Function.

Main algorithm of Tupper.

Break region into square regions for each square region, subdivide into 4 regions. For each check if there is no solution, if so add to white. Otherwise, refine the region by repeating until we can't subdivide

At that point, check each pixel-by-pixel region for possible values.

Return red, black and white vectors of Regions.

ImplicitEquations.Ge

ImplicitEquations.Gt

ImplicitEquations.I_ — Function.

Screen a value using NaN values. Use as with f(x,y) = x*y * screen(x > 0)

Also aliased to I_(x>0)

An expression like x::OInterval > 0 is not Boolean, but rather a BInterval which allows for a "MAYBE" state. As such, a simple ternary operator, like x > 0 ? 1 : NaN won't work, to screen values.

ImplicitEquations.ImplicitEquations

ImplicitEquations.Le

ImplicitEquations.Lt

ImplicitEquations.MAYBE

ImplicitEquations.Neq

ImplicitEquations.OInterval

ImplicitEquations.Pred — Type.

A predicate is defined in terms of a function of two variables, an inquality, and either another function or a real number. They are conveniently created by the functions Lt, Le, Eq, Neq, Ge, and Gt. The equivalent unicode operators:

≪ (\ll[tab]),

≦ (\leqq[tab]),

⩵ (\Equal[tab]),

≶ (\lessgtr[tab]) or ≷ (\gtrless[tab]),

≧ (\geqq[tab]),

≫ (\leqq[tab]) may also be used.

The use of Julia's usual comparison operators is no longer supported.

To combine predicates, & and | can be used.

To negate a predicate, ! is used.

ImplicitEquations.Predicate

ImplicitEquations.Preds — Type.

Predicates can be joined together with either & or |. Individual predicates can be negated with !. The parsing rules require the individual predicates to be enclosed with parentheses, as in (f==0) | (g==0).

ImplicitEquations.RefinePixels

ImplicitEquations.Region

ImplicitEquations.TRUE

ImplicitEquations.asciigraph — Function.

Function to graph, using ascii characters, a predicate related to a two-dimensional equation, such as f == 0 or f < g, where f and g are functions, e.g., f(x,y) = y - sqrt(x). The positional arguments are L, R, B, T which indicate the x-y range of the graph of the equation. The keywork arguments W and H indicate the number of pixels to use.

Set offset=0 to see algorithm

A pixel is important, as the graph will color a pixel

"white" (e.g., " ") if no solution exists in the region indicated by the pixel

"black" (e.g., "x") if a solution is known to exist in the region

"red" (e.g. ".") if a solution might exist

ImplicitEquations.break_into_squares

ImplicitEquations.check_continuity

ImplicitEquations.check_inequality — Function.

Does this function have a value in the pixel satisfying the inequality? Return TRUE or MAYBE.

ImplicitEquations.compute — Function.

Compute whether predicate holds in a given region. Returns FALSE, TRUE or MAYBE

ImplicitEquations.compute_fxy

ImplicitEquations.cross_zero — Function.

Does this function have a zero crossing? Heuristic check.

We return TRUE or MAYBE. However, that leaves some functions showing too much red in the case where there is no zero.

ImplicitEquations.eval

ImplicitEquations.fill_bottom

ImplicitEquations.get_xs_ys

ImplicitEquations.imag_map

ImplicitEquations.include

ImplicitEquations.irange

ImplicitEquations.linterp

ImplicitEquations.negate_op

ImplicitEquations.plot_implicit — Constant.

Plotting of implicit functions.

An implicit function or equation is defined in terms of a logical condition and a function of two variables. These produce Predicate objects.

Predicates may be plotted over a specified region, the default begin [-5,5] x [-5,5].

The algorithm, breaks the region into blocks. The ultimate resolution is given by W=2^n and H=2^m. The algorithm used, due to Tupper, colors region if it knows the predicate is true or false, and otherwise resolves the region into 4 equal-sized subregions and test each subregion again to determine if it is true, false, or still maybe. This repeats until W and H can't be subdivided.

The plotting of a predicate simply plots each block that is knows satisfies the predicate "black" and each ambiguous block "red." By taking m and n larger the graphs look less "blocky" but take more time to render.

For text-based plots, the asciigraph function is available.

Examples:

using Plots, ImplicitEquations
pyplot()

a,b = -1,2
f(x,y) = y^4 - x^4 + a*y^2 + b*x^2
plot(f == 0)

## trident of Newton
c,d,e,h = 1,1,1,1
f(x,y) = x*y
g(x,y) =c*x^3 + d*x^2 + e*x + h
plot(eq(f,g), title="Trident of Newton") ## aka f ⩵ g (using Unicode\Equal[tab])

## inequality
f(x,y)= (y-5)*cos(4*sqrt((x-4)^2 + y^2))
g(x,y) = x*sin(2*sqrt(x^2 + y^2))
r = f < g
plot(r, (-10, 10), (-10, 10), N=9, M=9)  # (xmin, xmax), (ymin, ymax),

ImplicitEquations.preds

ImplicitEquations.screen — Function.

Screen a value using NaN values. Use as with f(x,y) = x*y * screen(x > 0)

Also aliased to I_(x>0)

An expression like x::OInterval > 0 is not Boolean, but rather a BInterval which allows for a "MAYBE" state. As such, a simple ternary operator, like x > 0 ? 1 : NaN won't work, to screen values.

ImplicitEquations.xy_region

ImplicitEquations.xyrange

ImplicitEquations.≦

ImplicitEquations.≧

ImplicitEquations.≪

ImplicitEquations.≫

ImplicitEquations.≶

ImplicitEquations.≷

ImplicitEquations.⩵