Docstrings

Proj4.Cdouble15

Proj4.Cdouble21

Proj4.Cdouble6

Proj4.Proj4

Proj4.ProjUV

Proj4.Projection — Type.

Cartographic projection type

Proj4.__init__

Proj4._compare_datums — Function.

Returns true if the two datums are identical, otherwise false.

Proj4._errno — Function.

Get error number

Proj4._free — Function.

Free C datastructure associated with a projection. For internal use!

Proj4._fwd! — Function.

forward projection from Lat/Lon to X/Y (only supports 2 dimensions)

Row-wise projection from Lat/Lon to X/Y (only supports 2 dimensions)

Proj4._geod

Proj4._geod_direct! — Function.

Solve the direct geodesic problem.

Args:

g        - the geod_geodesic object specifying the ellipsoid.
lonlat   - where lat ∈ [-90, 90], lon ∈ [-540, 540), modified in-place to [dest] (described below)
azimuth  - azimuth (degrees) ∈ [-540, 540)
distance - distance (metres) to move from (lon,lat); can be negative

Returns:

dest     - destination after moving for [distance] metres in [azimuth] direction.
azi      - forward azimuth (degrees) at destination [dest].

Remarks:

If either point is at a pole, the azimuth is defined by keeping the longitude fixed,
writing lat = 90 +/- eps, and taking the limit as eps -> 0+. An arc length greater than 180deg
signifies a geodesic which is not a shortest path.

Proj4._geod_inverse — Function.

Solve the inverse geodesic problem.

Args:

g       - the geod_geodesic object specifying the ellipsoid.
lonlat1 - point 1 (degrees), where lat ∈ [-90, 90], lon ∈ [-540, 540) 
lonlat2 - point 2 (degrees), where lat ∈ [-90, 90], lon ∈ [-540, 540)

Returns:

dist    - distance between point 1 and point 2 (meters).
azi1    - azimuth at point 1 (degrees) ∈ [-180, 180)
azi2    - (forward) azimuth at point 2 (degrees) ∈ [-180, 180)

Remarks:

If either point is at a pole, the azimuth is defined by keeping the longitude fixed,
writing lat = 90 +/- eps, and taking the limit as eps -> 0+.

Proj4._geodesic

Proj4._get_def — Function.

Get projection definition string in the proj.4 plus format

Proj4._get_release — Function.

Get a string describing the underlying version of libproj in use

Proj4._get_spheroid_defn — Function.

Fetch the internal definition of the spheroid as a tuple (a, es), where

a = major_axis
es = eccentricity squared

Proj4._init_plus

Proj4._inv! — Function.

inverse projection from X/Y to Lat/Lon (only supports 2 dimensions)

Row-wise projection from X/Y to Lat/Lon (only supports 2 dimensions)

Proj4._is_geocent

Proj4._is_latlong

Proj4._strerrno — Function.

Get human readable error string from proj.4 error code

Get global errno string in human readable form

Proj4._transform! — Function.

Low level interface to libproj transform, C_NULL can be passed in for z, if it's 2-dimensional

Proj4._version

Proj4.ch_path

Proj4.check_deps

Proj4.compare_datums — Function.

Returns true if the datums for the two projections are identical

Proj4.depsjl_path

Proj4.epsg

Proj4.epsg_path

Proj4.error_message — Function.

Get a global error string in human readable form

Proj4.esri

Proj4.esri_extra_path

Proj4.esri_path

Proj4.eval

Proj4.freeProjection

Proj4.geod_destination

Proj4.geod_destination!

Proj4.geod_direct

Proj4.geod_direct!

Proj4.geod_distance

Proj4.geod_geodesic — Type.

Construct an ellipsoid of revolution with

equatorial radius a,
flattening f,

Remark, we construct it when constructing a [Projection] using the formulas:

f = (a − b)/a = 1 − sqrt(1-es)
n = (a − b)/(a + b) = f /(2 − f)
es = (a^2 − b^2)/a^2 = f(2 − f)

Reference: equations (1)-(3) of Algorithms for geodesics (arXiv:1109.4448v2 [physics.geo-ph] 28 Mar 2012)

Proj4.geod_inverse

Proj4.gl27_path

Proj4.has_geodesic_support

Proj4.ignf_path

Proj4.include

Proj4.is_geocent — Function.

Return true if the projection is a geocentric coordinate system

Proj4.is_latlong — Function.

Return true if the projection is a geographic coordinate system

Proj4.libproj

Proj4.lonlat2xy — Function.

Returns the forward projection from LonLat to XY in the given projection (only supports 2 dimensions)

Proj4.lonlat2xy! — Function.

Returns the forward projection from LatLon to XY in the given projection, modifying the input lonlat inplace (only supports 2 dimensions)

Proj4.nad27_path

Proj4.nad83_path

Proj4.nad_lst_path

Proj4.null_geodesic

Proj4.other_extra_path

Proj4.proj_def_dat_path

Proj4.spheroid_params — Function.

Return the definition of the spheroid as a tuple (a, es), where

a = major_axis
es = eccentricity squared

Proj4.transform — Function.

transform(src_projection, dest_projection, position [, radians=false])

Transform between geographic or projected coordinate systems, returning the transformed points in a Float64 array the same shape as position.

Proj4.transform! — Function.

transform!(src_projection, dest_projection, position [, radians=false])

Transform between geographic or projected coordinate systems, modifying position in place.

Args:

src      - Source coordinate system definition
dest     - Destination coordinate system definition
position - An array of coordinates to be transformed in place.  If `position` is a
           Vector of length 2 or 3 it's treated as a single point.  For
           geographic coordinate systems, the first two columns are the
           *longitude* and *latitude*, in that order.  To transform an
           array of points, a matrix of shape Nx2 or Nx3 may be used.
radians  - If true, treat geographic lon,lat coordinates as radians on
           input and output.

Returns:

position - Transformed position

Proj4.version — Constant.

Parsed version number for the underlying version of libproj

Proj4.world_path

Proj4.xy2lonlat — Function.

Returns the inverse projection from XY to LatLon in the given projection (only supports 2 dimensions)

Proj4.xy2lonlat! — Function.

Returns the inverse projection from XY to LonLat in the given projection, modifying the input xy inplace (only supports 2 dimensions)