Docstrings

GaussQuadrature.EndPt

GaussQuadrature.GaussQuadrature

GaussQuadrature.both

GaussQuadrature.chebyshev — Function.

x, w = chebyshev(T, n, kind=1, endpt=neither)

Returns points x and weights w for the n-point Gauss-Chebyshev rule for the interval -1 < x < 1 with weight function

w(x) = 1 / sqrt(1-x²)   if kind=1
w(x) = sqrt(1-x²)       if kind=2.

Use endpt=left, right or both for the left Radau, right Radau or Lobatto rules, respectively.

Convenience function with type T = Float64:

x, w = chebyshev(n, kind=1, endpt=neither)

GaussQuadrature.chebyshev_coefs

GaussQuadrature.custom_gauss_rule — Function.

x, w = custom_gauss_rule(lo, hi, a, b, endpt, maxits=maxiterations[T])

Generates the points x and weights w for a Gauss rule with weight function w(x) on the interval lo < x < hi.

The arrays a and b hold the coefficients (as given, for instance, by legendre_coeff) in the three-term recurrence relation for the monic orthogonal polynomials p(0,x), p(1,x), p(2,x), ... , that is,

p(k, x) = (x-a[k]) p(k-1, x) - b[k]² p(k-2, x),    k ≥ 1,

where p(0, x) = 1 and, by convention, p(-1, x) = 0 with

          hi
b[1]^2 = ∫  w(x) dx.
         lo

Thus, p(k, x) = xᵏ + lower degree terms and

 hi
∫  p(j, x) p(k, x) w(x) dx = 0 if j ≠ k.
lo

GaussQuadrature.eval

GaussQuadrature.hermite — Function.

x, w = hermite(T, n)

Returns points x and weights w for the n-point Gauss-Laguerre rule for the interval -∞ < x < ∞ with weight function

w(x) = exp(-x²).

Convenience function with type T = Float64:

x, w = hermite(n)

GaussQuadrature.hermite_coefs

GaussQuadrature.include

GaussQuadrature.jacobi — Function.

x, w = jacobi(n, α, β, endpt=neither)

Returns points x and weights w for the n-point Gauss-Jacobi rule for the interval -1 < x < 1 with weight function

w(x) = (1-x)ᵅ (1+x)ᵝ.

Use endpt=left, right or both for the left Radau, right Radau or Lobatto rules, respectively.

GaussQuadrature.jacobi_coefs

GaussQuadrature.laguerre — Function.

x, w = laguerre(n, α, endpt=neither)

Returns points x and weights w for the n-point Gauss-Laguerre rule for the interval 0 < x < ∞ with weight function

w(x) = xᵅ exp(-x),   α > -1.

Use endpt=left for the left Radau rule.

GaussQuadrature.laguerre_coefs

GaussQuadrature.left

GaussQuadrature.legendre — Function.

x, w = legendre(T, n, endpt=neither)

Returns points x and weights w for the n-point Gauss-Legendre rule for the interval -1 < x < 1 with weight function w(x) = 1.

Use endpt=left, right or both for the left Radau, right Radau or Lobatto rules, respectively.

Convenience function with type T = Float64:

x, w = legendre(n, endpt=neither)

GaussQuadrature.legendre_coefs

GaussQuadrature.logweight — Function.

x, w = logweight(T, n, r=0, endpt=neither)

Returns points x and weights w for the n-point Gauss rule on the interval 0 < x < 1 with weight function

w(x) = xʳ log(1/x),    r ≥ 0.

Convenience function with type T = Float64:

x, w = logweight(n, r=0, endpt=neither)

More general method works when w(x) = x^ρ log(1/x) for real ρ > -1:

x, w = logweight(n, ρ, endpt=neither)

GaussQuadrature.logweight_coefs

GaussQuadrature.maxiterations

GaussQuadrature.modified_chebyshev — Function.

α, β, σ = modified_chebyshev(a, b, ν)

Implements the modified Chebyshev algorithm described in doc/notes.tex and used in logweight_coefs.

GaussQuadrature.modified_moments

GaussQuadrature.neither

GaussQuadrature.orthonormal_poly

GaussQuadrature.right

GaussQuadrature.shifted_legendre_coefs

GaussQuadrature.solve

GaussQuadrature.special_eigenproblem!