Docstrings

ReusableFunctions.ReusableFunctions — Module.

MADS: Model Analysis & Decision Support in Julia (Mads.jl v1.0) 2016

http://mads.lanl.gov http://madsjulia.lanl.gov http://gitlab.com/mads/Mads.jl

Licensing: GPLv3: http://www.gnu.org/licenses/gpl-3.0.html

Copyright 2016. Los Alamos National Security, LLC. All rights reserved.

This material was produced under U.S. Government contract DE-AC52-06NA25396 for Los Alamos National Laboratory, which is operated by Los Alamos National Security, LLC for the U.S. Department of Energy. The Government is granted for itself and others acting on its behalf a paid-up, nonexclusive, irrevocable worldwide license in this material to reproduce, prepare derivative works, and perform publicly and display publicly. Beginning five (5) years after –––––––- November 17, 2015, –––––––––––––––––––––––––––––––– subject to additional five-year worldwide renewals, the Government is granted for itself and others acting on its behalf a paid-up, nonexclusive, irrevocable worldwide license in this material to reproduce, prepare derivative works, distribute copies to the public, perform publicly and display publicly, and to permit others to do so.

NEITHER THE UNITED STATES NOR THE UNITED STATES DEPARTMENT OF ENERGY, NOR LOS ALAMOS NATIONAL SECURITY, LLC, NOR ANY OF THEIR EMPLOYEES, MAKES ANY WARRANTY, EXPRESS OR IMPLIED, OR ASSUMES ANY LEGAL LIABILITY OR RESPONSIBILITY FOR THE ACCURACY, COMPLETENESS, OR USEFULNESS OF ANY INFORMATION, APPARATUS, PRODUCT, OR PROCESS DISCLOSED, OR REPRESENTS THAT ITS USE WOULD NOT INFRINGE PRIVATELY OWNED RIGHTS.

LA-CC-15-080; Copyright Number Assigned: C16008

ReusableFunctions.checkfunctionignore_keywords

ReusableFunctions.checkfunctionkeywords

ReusableFunctions.checkhashfilename — Function.

Check if a file with a filename based on hash exists

ReusableFunctions.computes

ReusableFunctions.eval

ReusableFunctions.getfunctionkeywords

ReusableFunctions.gethashfilename — Function.

Define a filename based on hash

ReusableFunctions.include

ReusableFunctions.loadresultfile — Function.

Load JLD result file

ReusableFunctions.maker3function — Function.

Make a reusable function expecting both regular and keyword arguments

ReusableFunctions.quiet

ReusableFunctions.quietoff — Function.

Make ReusableFunctions not quiet

ReusableFunctions.quieton — Function.

Make ReusableFunctions quiet

ReusableFunctions.resetcomputes — Function.

Reset computes counter

ReusableFunctions.resetrestarts — Function.

Reset restarts counter

ReusableFunctions.restarts

ReusableFunctions.saveresultfile — Function.

Save JLD result file