Docstrings

RigidBodyTreeInspector.DrakeVisualizerSink

RigidBodyTreeInspector.InterpolatableArray — Type.

Interpolations.jl requires that one(::Type{T}) be defined for any data type we want to interpolate. Rather than defining one(::Type{Vector}) here, which might have unforeseen consequences in other packages, we'll create a very simple wrapper type that just knows one() and *

RigidBodyTreeInspector.Manipulate

RigidBodyTreeInspector.RigidBodyTreeInspector

RigidBodyTreeInspector.Skeleton

RigidBodyTreeInspector.URDFVisuals

DrakeVisualizer.Visualizer — Type.

Construct a DrakeVisualizer.Visualizer for the given mechanism by constructing simple geometries just from the structure of the kinematic tree. If show_intertias is true, then also construct equivalent inertial ellipsoids for every link.

RigidBodyTreeInspector.addgeometry!

RigidBodyTreeInspector.animate — Function.

animate(vis::Visualizer, mechanism::Mechanism{Float64},
        times::Vector{Float64},
        configurations::Vector{Vector{Float64}};
        fps::Float64=60, realtimerate::Float64=1.)

Animate the given mechanism passing through a time-coded series of configurations by linearly interpolating the configuration vectors.

RigidBodyTreeInspector.apply_scaling

RigidBodyTreeInspector.create_geometry

RigidBodyTreeInspector.eval

RigidBodyTreeInspector.include

RigidBodyTreeInspector.inspect

RigidBodyTreeInspector.inspect!

RigidBodyTreeInspector.manipulate

RigidBodyTreeInspector.normalize_configuration!

RigidBodyTreeInspector.parse_urdf

RigidBodyTreeInspector.rbd

RigidBodyTreeInspector.remove_scaling — Function.

The remote tree viewer interface has no notion of scale (only a quaternion and translation), so we have to do some monkey business to push the scaling into the geometry itself.

Currently, this only works for converting spheres into ellipsoids and assumes the scaling is positive.

RigidBodyTreeInspector.setgeometry!

RigidBodyTreeInspector.settransform!

RigidBodyTreeInspector.unique_frame_name

RigidBodyTreeInspector.visual_elements