Docstrings

WCS.@check_prop

WCS.@check_type

WCS.@same_size

WCS.HDR_ALL

WCS.HDR_ALLIMG

WCS.HDR_AUXIMG

WCS.HDR_BIMGARR

WCS.HDR_CD00i00j

WCS.HDR_CD0i_0ja

WCS.HDR_CNAMn

WCS.HDR_CROTAia

WCS.HDR_DOBSn

WCS.HDR_EPOCHa

WCS.HDR_IMGHEAD

WCS.HDR_LONGKEY

WCS.HDR_NONE

WCS.HDR_PC00i00j

WCS.HDR_PC0i_0ja

WCS.HDR_PIXLIST

WCS.HDR_PROJPn

WCS.HDR_PS0i_0ma

WCS.HDR_PV0i_0ma

WCS.HDR_RADECSYS

WCS.HDR_REJECT

WCS.HDR_STRICT

WCS.HDR_VELREFa

WCS.HDR_VSOURCE

WCS.PSCard

WCS.PVCard

WCS.WCS

WCS.WCSErr

WCS.WCSTransform — Type.

WCSTransform(naxis; kvs...)

Construct a WCS transformation with the given number of axes naxis. Keyword arguments can be passed to set various attributes of the transform. Specifying keyword arguments is equivalent to setting them after construction:

julia> wcs = WCSTransform(2; crpix=[1000., 1000.])

is equilvalent to:

julia> wcs = WCSTransform(2)

julia> wcs.crpix = [1000., 1000.]

WCS.assert_ok

WCS.celprm

WCS.check_deps

WCS.convert_string

WCS.eval

WCS.free!

WCS.from_header — Function.

fromheader(header[; relax=WCS.HDRALL, ctrl=0, ignore_rejected=false, table=false])

Parse the FITS image header in the String header, returning a Vector{WCSTransform} giving all the transforms defined in the header. The relax determines the treatment of non-standard keywords. The default is to accept all known non-standard keywords. Use relax=WCS.HDR_NONE to ignore all non-standard keywords. Use, e.g., relax=(WCS.HDR_RADECSYS & WCS.HDR_CROTAia) to only accept selected non-standard keywords.

WCS.get_error_message

WCS.include

WCS.libwcs

WCS.linprm

WCS.pix_to_world — Function.

pixtoworld(wcs, pixcoords)

Convert the array of pixel coordinates pixcoords to world coordinates according to the WCSTransform wcs. pixcoords should be a 2-d array where "pixcoords[:, i]" is the i-th set of coordinates, or a 1-d array representing a single set of coordinates.

The return value is the same shape as pixcoords.

WCS.pix_to_world! — Function.

pixtoworld!(wcs, pixcoords, worldcoords[; stat=, imcoords=, phi=, theta=])

Convert the array of pixel coordinates pixcoords to world coordinates according to the WCSTransform wcs, storing the result in the worldcoords and stat arrays. pixcoords should be a 2-d array where "pixcoords[:, i]" is the i-th set of coordinates, or a 1-d array representing a single set of coordinates. worldcoords must be the same size and type as pixcoords.

If given, the arrays stat, imcoords, phi, theta will be used to store intermediate results. Their sizes and types must all match pixcoords, except for stat which should be the same size but of type Cint (typically Int32).

WCS.prjprm

WCS.spcprm

WCS.to_header — Function.

toheader(wcs[; relax=WCS.HDRNONE])

Encode the WCSTransform wcs as a FITS header string. The relax keyword controls how non-standard extensions to the WCS standard are handled.

WCS.unsafe_store_vec!

WCS.wcs_lock

WCS.wcslib_version

WCS.world_to_pix — Function.

worldtopix(wcs, worldcoords)

Convert the array of world coordinates worldcoords to pixel coordinates according to the WCSTransform wcs. worldcoords is a 2-d array where "worldcoords[:, i]" is the i-th set of coordinates, or a 1-d array representing a single set of coordinates.

The return value is the same size as worldcoords.

WCS.world_to_pix! — Function.

worldtopix!(wcs, worldcoords, pixcoords[; stat=, phi=, theta=, imcoords=])

Convert the array of pixel coordinates worldcoords to pixel coordinates according to the WCSTransform wcs, storing the result in the pixcoords array. worldcoords should be a 2-d array where "worldcoords[:, i]" is the i-th set of coordinates, or a 1-d array representing a single set of coordinates. pixcoords must be the same size and type as worldcoords.

If given, the arrays stat, imcoords, phi, theta will be used to store intermediate results. Their sizes and types must all match worldcoords, except for stat which should be the same size but of type Cint (typically Int32).