Docstrings

ControlSystems.AbstractSimulator

ControlSystems.AbstractSystem

ControlSystems.Bodeplot

ControlSystems.ControlSystems

ControlSystems.Impulseplot

ControlSystems.LQG — Type.

G = LQG(A,B,C,D, Q1, Q2, R1, R2; qQ=0, qR=0, integrator=false)
G = LQG(sys, args...; kwargs...)

Return an LQG object that describes the closed control loop around the process sys=ss(A,B,C,D) where the controller is of LQG-type. The controller is specified by weight matrices R1,R2 that penalizes state deviations and control signal variance respectively, and covariance matrices Q1,Q2 which specify state drift and measurement covariance respectively. This constructor calls lqr and kalman and forms the closed-loop system.

If integrator=true, the resulting controller will have intregral action. This is achieved by adding a model of a constant disturbance on the inputs to the system described by A,B,C,D.

qQ and qR can be set to incorporate loop transfer recovery, i.e.,

L = lqr(A, B, Q1+qQ*C'C, Q2)
K = kalman(A, C, R1+qR*B*B', R2)

Fields

When the LQG-object is populated by the lqg-function, the following fields have been made available

L is the feedback matrix, such that A-BL is stable. Note that the length of the state vector (and the width of L) is increased by the number of inputs if the option integrator=true.

K is the kalman gain such that A-KC is stable

sysc is a dynamical system describing the controller u=L*inv(A-BL-KC+KDL)Ky

Functions

Several other properties of the object are accessible with the indexing function getindex() and are called with the syntax G[:function]. The available functions are (some have many alternative names, separated with / )

-G[:cl] / G[:closedloop] is the closed-loop system, including observer, from reference to output, precompensated to have static gain 1 (u = −Lx + lᵣr). -G[:S] / G[:Sin] Input sensitivity function -G[:T] / G[:Tin] Input complementary sensitivity function -G[:Sout] Output sensitivity function -G[:Tout] Output complementary sensitivity function -G[:CS] The transfer function from measurement noise to control signal -G[:DS] The transfer function from input load disturbance to output -G[:lt] / G[:looptransfer] / G[:loopgain] = PC -G[:rd] / G[:returndifference] = I + PC -G[:sr] / G[:stabilityrobustness] = I + inv(PC) -G[:sysc] / G[:controller] Returns the controller as a StateSpace-system

It is also possible to access all fileds using the G[:symbol] syntax, the fields are P ,Q1,Q2,R1,R2,qQ,qR,sysc,L,K,integrator

Example

eye(n) = Matrix{Float64}(I,n,n) # For convinience

qQ = 1
qR = 1
Q1 = 10eye(4)
Q2 = 1eye(2)
R1 = 1eye(6)
R2 = 1eye(2)

G = LQG(sys, Q1, Q2, R1, R2, qQ=qQ, qR=qR, integrator=true)

Gcl = G[:cl]
T = G[:T]
S = G[:S]
sigmaplot([S,T],exp10.(range(-3, stop=3, length=1000)))
stepplot(Gcl)

ControlSystems.LTISystem

ControlSystems.Lsimplot

ControlSystems.Nicholsplot

ControlSystems.Nyquistplot

ControlSystems.Plots

ControlSystems.Pzmap

ControlSystems.Rlocusplot

ControlSystems.Sigmaplot

ControlSystems.Simulator — Type.

Simulator

Fields:

P::StateSpace
f = (x,p,t) -> x
y = (x,t)   -> y

ControlSystems.SisoRational

ControlSystems.SisoTf

ControlSystems.SisoZpk

ControlSystems.StateSpace

ControlSystems.Stepplot

ControlSystems.TransferFunction — Type.

F(s), F(omega, true), F(z, false)

Notation for frequency response evaluation.

F(s) evaluates the continuous-time transfer function F at s.

F(omega,true) evaluates the discrete-time transfer function F at exp(iTsomega)

F(z,false) evaluates the discrete-time transfer function F at z

ControlSystems._LQG

ControlSystems._LQGi

ControlSystems._PlotScale

ControlSystems._PlotScaleFunc

ControlSystems._PlotScaleStr

ControlSystems.__CONTROLSYSTEMS_SOURCE_DIR__

ControlSystems._allGainCrossings

ControlSystems._allPhaseCrossings

ControlSystems._bounds_and_features

ControlSystems._check_consistent_sampling_time

ControlSystems._default_Ts

ControlSystems._default_freq_vector

ControlSystems._default_time_data

ControlSystems._default_time_vector

ControlSystems._findCrossings

ControlSystems._fix_conjugate_pairs! — Function.

Typically LAPACK returns a vector with, e.g., eigenvalues to a real matrix, they are paired up in exact conjugate pairs. This fact is used in some places for working with zpk representations of LTI systems. eigvals(A) returns a on this form, however, for generalized eigenvalues there is a small numerical discrepancy, which breaks some functions. This function fixes small discrepancies in the conjugate pairs.

ControlSystems._issmooth

ControlSystems._preprocess_for_freqresp

ControlSystems._printcoefficient — Function.

Heurisitc function that tries to add parentheses when printing the coeffcient for systems on zpk form. Should at least handle the following types Measurment, Dual, Sym.

ControlSystems._same_io_dims

ControlSystems._string_mat_with_headers

ControlSystems.acker

ControlSystems.append — Function.

append(systems::StateSpace...), append(systems::TransferFunction...)

Append systems in block diagonal form

ControlSystems.balance — Function.

T, B = balance(A[, perm=true])

Compute a similarity transform T resulting in B = T\A*T such that the row and column norms of B are approximately equivalent. If perm=false, the transformation will only scale, and not permute A.

ControlSystems.balance_statespace — Function.

A, B, C, T = balance_statespace{S}(A::Matrix{S}, B::Matrix{S}, C::Matrix{S}, perm::Bool=false)

sys, T = balance_statespace(sys::StateSpace, perm::Bool=false)

Computes a balancing transformation T that attempts to scale the system so that the row and column norms of [TA/T TB; C/T 0] are approximately equal. If perm=true, the states in A are allowed to be reordered.

This is not the same as finding a balanced realization with equal and diagonal observability and reachability gramians, see balreal

ControlSystems.balance_transform — Function.

T = balance_transform{R}(A::Matrix{R}, B::Matrix{R}, C::Matrix{R}, perm::Bool=false)

T = balance_transform(sys::StateSpace, perm::Bool=false) = balance_transform(A,B,C,perm)

Computes a balancing transformation T that attempts to scale the system so that the row and column norms of [TA/T TB; C/T 0] are approximately equal. If perm=true, the states in A are allowed to be reordered.

No balancing will be done if A, B C are not BLAS compatible

This is not the same as finding a balanced realization with equal and diagonal observability and reachability gramians, see balreal See also balance_statespace, balance

ControlSystems.balreal — Function.

sysr, G = balreal(sys::StateSpace)

Calculates a balanced realization of the system sys, such that the observability and reachability gramians of the balanced system are equal and diagonal G

See also gram, baltrunc

Glad, Ljung, Reglerteori: Flervariabla och Olinjära metoder

ControlSystems.baltrunc — Function.

sysr, G = baltrunc(sys::StateSpace, atol = √ϵ, rtol=1e-3, unitgain=true)

Reduces the state dimension by calculating a balanced realization of the system sys, such that the observability and reachability gramians of the balanced system are equal and diagonal G, and truncating it such that all states corresponding to singular values less than atol and less that rtol σmax are removed. If unitgain=true, the matrix D is chosen such that unit static gain is achieved.

See also gram, balreal

Glad, Ljung, Reglerteori: Flervariabla och Olinjära metoder

ControlSystems.blockdiag

ControlSystems.bode — Function.

mag, phase, w = bode(sys[, w])

Compute the magnitude and phase parts of the frequency response of system sys at frequencies w

mag and phase has size (length(w), ny, nu)

ControlSystems.bodeplot — Function.

fig = bodeplot(sys, args...), bodeplot(LTISystem[sys1, sys2...], args...; plotphase=true, kwargs...)

Create a Bode plot of the LTISystem(s). A frequency vector w can be optionally provided.

kwargs is sent as argument to Plots.plot.

ControlSystems.bodeplot!

ControlSystems.c2d — Function.

[sysd, x0map] = c2d(sys, Ts, method=:zoh)

Convert the continuous system sys into a discrete system with sample time Ts, using the provided method. Currently only :zoh and :foh are provided.

Returns the discrete system sysd, and a matrix x0map that transforms the initial conditions to the discrete domain by x0_discrete = x0map*[x0; u0]

ControlSystems.c2d_poly2poly — Function.

c2d_poly2poly(ro,h)

returns the polynomial coefficients in discrete time given polynomial coefficients in continuous time

ControlSystems.c2d_roots2poly — Function.

c2d_roots2poly(ro,h)

returns the polynomial coefficients in discrete time given a vector of roots in continuous time

ControlSystems.care — Function.

care(A, B, Q, R)

Compute 'X', the solution to the continuous-time algebraic Riccati equation, defined as A'X + XA - (XB)R^-1(B'X) + Q = 0, where R is non-singular.

Algorithm taken from: Laub, "A Schur Method for Solving Algebraic Riccati Equations." http://dspace.mit.edu/bitstream/handle/1721.1/1301/R-0859-05666488.pdf

ControlSystems.charpoly

ControlSystems.check_real — Function.

If TR is Complex and T is Real, check that every pole is matched to its conjugate this assumes that the compelx poles are ordered as they are output by the LAPACK routines that return complex-conjugated values, i.e., (x+iy) is followed by (x-iy)

ControlSystems.covar — Function.

P = covar(sys, W)

Calculate the stationary covariance P = E[y(t)y(t)'] of an lti-model sys, driven by gaussian white noise 'w' of covariance E[w(t)w(τ)]=W*δ(t-τ) where δ is the dirac delta.

The ouput is if Inf if the system is unstable. Passing white noise directly to the output will result in infinite covariance in the corresponding outputs (DWD' .!= 0) for contunuous systems.

ControlSystems.cswap!

ControlSystems.ctrb — Function.

ctrb(A, B) or ctrb(sys)

Compute the controllability matrix for the system described by (A, B) or sys.

Note that checking for controllability by computing the rank from ctrb is not the most numerically accurate way, a better method is checking if gram(sys, :c) is positive definite.

ControlSystems.dab — Function.

DAB Solves the Diophantine-Aryabhatta-Bezout identity

X,Y = DAB(A,B,C)

AX + BY = C, where A, B, C, X and Y are polynomials and deg Y = deg A - 1.

See Computer-Controlled Systems: Theory and Design, Third Edition Karl Johan Åström, Björn Wittenmark

ControlSystems.damp — Function.

Wn, zeta, ps = damp(sys)

Compute the natural frequencies, Wn, and damping ratios, zeta, of the poles, ps, of sys

ControlSystems.dampreport — Function.

dampreport(sys)

Display a report of the poles, damping ratio, natural frequency, and time constant of the system sys

ControlSystems.dare — Function.

dare(A, B, Q, R)

Compute X, the solution to the discrete-time algebraic Riccati equation, defined as A'XA - X - (A'XB)(B'XB + R)^-1(B'XA) + Q = 0, where A and R are non-singular.

Algorithm taken from: Laub, "A Schur Method for Solving Algebraic Riccati Equations." http://dspace.mit.edu/bitstream/handle/1721.1/1301/R-0859-05666488.pdf

ControlSystems.dcgain — Function.

dcgain(sys)

Compute the dcgain of system sys.

equal to G(0) for continuous-time systems and G(1) for discrete-time systems.

ControlSystems.delaymargin — Function.

dₘ = delaymargin(G::LTISystem)

Only supports SISO systems

ControlSystems.den

ControlSystems.denpoly

ControlSystems.denvec

ControlSystems.diagonalize — Function.

dsys = diagonalize(s::StateSpace, digits=12) Diagonalizes the system such that the A-matrix is diagonal.

ControlSystems.dkalman — Function.

dkalman(A, C, R1, R2) kalman(sys, R1, R2)`

Calculate the optimal Kalman gain for discrete time systems

ControlSystems.dlqr — Function.

dlqr(A, B, Q, R), dlqr(sys, Q, R)

Calculate the optimal gain matrix K for the state-feedback law u[k] = K*x[k] that minimizes the cost function:

J = sum(x'Qx + u'Ru, 0, inf).

For the discrte time model x[k+1] = Ax[k] + Bu[k].

See also lqg

Usage example:

using LinearAlgebra # For identity matrix I
h = 0.1
A = [1 h; 0 1]
B = [0;1]
C = [1 0]
sys = ss(A,B,C,0, h)
Q = I
R = I
L = dlqr(A,B,Q,R) # lqr(sys,Q,R) can also be used

u(t,x) = -L*x # Form control law,
t=0:h:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0)
plot(t,x, lab=["Position", "Velocity"]', xlabel="Time [s]")

ControlSystems.dlyap — Function.

dlyap(A, Q)

Compute the solution X to the discrete Lyapunov equation AXA' - X + Q = 0.

ControlSystems.eval

ControlSystems.evalfr — Function.

evalfr(sys, x) Evaluate the transfer function of the LTI system sys at the complex number s=x (continuous-time) or z=x (discrete-time).

For many values of x, use freqresp instead.

ControlSystems.fastrank

ControlSystems.feedback — Function.

feedback(L) Returns L/(1+L) feedback(P1,P2) Returns P1/(1+P1*P2)

feedback(sys)

feedback(sys1,sys2)

Forms the negative feedback interconnection

>-+ sys1 +-->
  |      |
 (-)sys2 +

If no second system is given, negative identity feedback is assumed

ControlSystems.feedback2dof — Function.

feedback2dof(P,R,S,T) Return BT/(AR+ST) where B and A are the numerator and denomenator polynomials of P respectively feedback2dof(B,A,R,S,T) Return BT/(AR+ST)

ControlSystems.freqresp — Function.

sys_fr = freqresp(sys, w)

Evaluate the frequency response of a linear system

w -> C*((iw*im -A)^-1)*B + D

of system sys over the frequency vector w.

ControlSystems.gangoffour — Function.

S,D,N,T = gangoffour(P,C), gangoffour(P::AbstractVector,C::AbstractVector)

Given a transfer function describing the Plant P and a transferfunction describing the controller C, computes the four transfer functions in the Gang-of-Four.

S = 1/(1+PC) Sensitivity function

D = P/(1+PC)

N = C/(1+PC)

T = PC/(1+PC) Complementary sensitivity function

Only supports SISO systems

ControlSystems.gangoffourplot — Function.

fig = gangoffourplot(P::LTISystem, C::LTISystem), gangoffourplot(P::Union{Vector, LTISystem}, C::Vector; plotphase=false)

Gang-of-Four plot.

kwargs is sent as argument to Plots.plot.

ControlSystems.gangofseven — Function.

S, D, N, T, RY, RU, RE = gangofseven(P,C,F)

Given transfer functions describing the Plant P, the controller C and a feed forward block F, computes the four transfer functions in the Gang-of-Four and the transferfunctions corresponding to the feed forward.

S = 1/(1+PC) Sensitivity function

D = P/(1+PC)

N = C/(1+PC)

T = PC/(1+PC) Complementary sensitivity function

RY = PCF/(1+PC)

RU = CF/(1+P*C)

RE = F/(1+P*C)

Only supports SISO systems

ControlSystems.getColorSys

ControlSystems.getLogTicks

ControlSystems.getStyleSys

ControlSystems.get_A

ControlSystems.get_B

ControlSystems.get_C

ControlSystems.get_D

ControlSystems.get_Ts

ControlSystems.getlims — Function.

Get atributes from xlims or ylims default to extrema(wmag) if xlims/ylims not defined or empty

ControlSystems.getpoles

ControlSystems.gram — Function.

gram(sys, opt)

Compute the grammian of system sys. If opt is :c, computes the controllability grammian. If opt is :o, computes the observability grammian.

ControlSystems.impulse — Function.

y, t, x = impulse(sys[, Tf]) or y, t, x = impulse(sys[, t])

Calculate the impulse response of system sys. If the final time Tf or time vector t is not provided, one is calculated based on the system pole locations.

y has size (length(t), ny, nu), x has size (length(t), nx, nu)

ControlSystems.impulseplot — Function.

impulseplot(sys[, Tf[,  Ts]])

Plot step response of sys with optional final time Tf and discretization time Ts. If not defined, suitable values are chosen based on sys.

ControlSystems.impulseplot!

ControlSystems.include

ControlSystems.index2range — Function.

outs = index2range(ind1, ind2) Helper function to convert indexes with scalars to ranges. Used to avoid dropping dimensions

ControlSystems.indirect_str

ControlSystems.interpolate

ControlSystems.iscontinuous — Function.

iscontinuous(sys)

Returns true if sys is continuous, else returns false.

ControlSystems.isproper — Function.

isproper(tf)

Returns true if the TransferFunction is proper. This means that order(den)

= order(num))

ControlSystems.issiso — Function.

issiso(sys)

Returns true if sys is SISO, else returns false.

ControlSystems.isstable — Function.

isstable(sys)

Returns true if sys is stable, else returns false.

ControlSystems.kalman — Function.

kalman(A, C, R1, R2) kalman(sys, R1, R2)`

Calculate the optimal Kalman gain

See also LQG

ControlSystems.laglink — Function.

laglink(a, M; h=0)

Returns a phase retarding link, the rule of thumb a = 0.1ωc guarantees less than 6 degrees phase margin loss. The bode curve will go from M, bend down at a/M and level out at 1 for frequencies > a

ControlSystems.leadlink — Function.

leadlink(b, N, K; h=0)

Returns a phase advancing link, the top of the phase curve is located at ω = b√(N) where the link amplification is K√(N) The bode curve will go from K, bend up at b and level out at KN for frequencies > bN

The phase advance at ω = b√(N) can be plotted as a function of N with leadlinkcurve()

Values of N < 1 will give a phase retarding link.

See also leadlinkat laglink

ControlSystems.leadlinkat — Function.

leadlinkat(ω, N, K; h=0)

Returns a phase advancing link, the top of the phase curve is located at ω where the link amplification is K√(N) The bode curve will go from K, bend up at ω/√(N) and level out at KN for frequencies > ω√(N)

The phase advance at ω can be plotted as a function of N with leadlinkcurve()

Values of N < 1 will give a phase retarding link.

See also leadlink laglink

ControlSystems.leadlinkcurve — Function.

Plot the phase advance as a function of N for a lead link (phase advance link)

If an input argument s is given, the curve is plotted from s to 10, else from 1 to 10.

See also Leadlink, leadlinkat

ControlSystems.loopshapingPI — Function.

kp,ki,C = loopshapingPI(P,ωp; ϕl,rl, phasemargin)

Selects the parameters of a PI-controller such that the Nyquist curve of P at the frequency ωp is moved to rl exp(i ϕl)

If phasemargin is supplied, ϕl is selected such that the curve is moved to an angle of phasemargin - 180 degrees

If no rl is given, the magnitude of the curve at ωp is kept the same and only the phase is affected, the same goes for ϕl if no phasemargin is given.

See also pidplots, stabregionPID

ControlSystems.lqg

ControlSystems.lqgi

ControlSystems.lqr — Function.

lqr(A, B, Q, R)

Calculate the optimal gain matrix K for the state-feedback law u = K*x that minimizes the cost function:

J = integral(x'Qx + u'Ru, 0, inf).

For the continuous time model dx = Ax + Bu.

lqr(sys, Q, R)

Solve the LQR problem for state-space system sys. Works for both discrete and continuous time systems.

See also LQG

Usage example:

using LinearAlgebra # For identity matrix I
A = [0 1; 0 0]
B = [0;1]
C = [1 0]
sys = ss(A,B,C,0)
Q = I
R = I
L = lqr(sys,Q,R)

u(t,x) = -L*x # Form control law,
t=0:0.1:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0)
plot(t,x, lab=["Position", "Velocity"]', xlabel="Time [s]")

ControlSystems.lsim — Function.

y, t, x = lsim(sys, u, t; x0, method])

y, t, x, uout = lsim(sys, u::Function, t; x0, method)

Calculate the time response of system sys to input u. If x0 is ommitted, a zero vector is used.

y, x, uout has time in the first dimension. Initial state x0 defaults to zero.

Continuous time systems are simulated using an ODE solver if u is a function. If u is an array, the system is discretized before simulation. For a lower level inteface, see ?Simulator and ?solve

u can be a function or a matrix/vector of precalculated control signals. If u is a function, then u(x,i) (u(x,t)) is called to calculate the control signal every iteration (time instance used by solver). This can be used to provide a control law such as state feedback u(x,t) = -L*x calculated by lqr. To simulate a unit step, use (x,i)-> 1, for a ramp, use (x,i)-> i*h, for a step at t=5, use (x,i)-> (i*h >= 5) etc.

Usage example:

using LinearAlgebra # For identity matrix I
A = [0 1; 0 0]
B = [0;1]
C = [1 0]
sys = ss(A,B,C,0)
Q = I
R = I
L = lqr(sys,Q,R)

u(x,t) = -L*x # Form control law,
t=0:0.1:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0)
plot(t,x, lab=["Position", "Velocity"]', xlabel="Time [s]")

ControlSystems.lsima — Function.

[y, t, x] = lsima(sys, t[, x0, method])

Calculate the time response of adaptive controller. If x0 is ommitted, a zero vector is used.

Continuous time systems are discretized before simulation. By default, the method is chosen based on the smoothness of the input signal. Optionally, the method parameter can be specified as either :zoh or :foh.

ControlSystems.lsimplot — Function.

fig = lsimplot(sys::LTISystem, u, t; x0=0, method)

lsimplot(LTISystem[sys1, sys2...], u, t; x0, method)

Calculate the time response of the LTISystem(s) to input u. If x0 is not specified, a zero vector is used.

Continuous time systems are discretized before simulation. By default, the method is chosen based on the smoothness of the input signal. Optionally, the method parameter can be specified as either :zoh or :foh.

ControlSystems.lsimplot!

ControlSystems.ltitr — Function.

ltitr(A, B, u[,x0])

ltitr(A, B, u::Function, iters[,x0])

Simulate the discrete time system x[k + 1] = A x[k] + B u[k], returning x. If x0 is not provided, a zero-vector is used.

If u is a function, then u(x,i) is called to calculate the control signal every iteration. This can be used to provide a control law such as state feedback u=-Lx calculated by lqr. In this case, an integrer iters must be provided that indicates the number of iterations.

ControlSystems.lyap

ControlSystems.margin — Function.

ωgₘ, gₘ, ωϕₘ, ϕₘ = margin{S<:Real}(sys::LTISystem, w::AbstractVector{S}; full=false, allMargins=false)

returns frequencies for gain margins, gain margins, frequencies for phase margins, phase margins

If !allMargins, return only the smallest margin

If full return also fullPhase

ControlSystems.marginplot — Function.

fig = marginplot(sys::LTISystem [,w::AbstractVector]; kwargs...), marginplot(sys::Vector{LTISystem}, w::AbstractVector; kwargs...)

Plot all the amplitude and phase margins of the system(s) sys. A frequency vector w can be optionally provided.

kwargs is sent as argument to Plots.plot.

ControlSystems.markovparam — Function.

markovparam(sys, n)

Compute the nth markov parameter of state-space system sys. This is defined as the following:

h(0) = D

h(n) = C*A^(n-1)*B

ControlSystems.minreal — Function.

tf = minreal(tf::TransferFunction, eps=sqrt(eps()))

Create a minimial representation of each transfer function in tf by cancelling poles and zeros will promote system to an appropriate numeric type

minsys = minreal(s::StateSpace, tol=sqrt(eps())) is implemented via baltrunc and returns a system on diagonal form.

ControlSystems.nicholsplot — Function.

fig = nicholsplot{T<:LTISystem}(systems::Vector{T}, w::AbstractVector; kwargs...)

Create a Nichols plot of the LTISystem(s). A frequency vector w can be optionally provided.

Keyword arguments:

text = true
Gains = [12, 6, 3, 1, 0.5, -0.5, -1, -3, -6, -10, -20, -40, -60]
pInc = 30
sat = 0.4
val = 0.85
fontsize = 10

pInc determines the increment in degrees between phase lines.

sat ∈ [0,1] determines the saturation of the gain lines

val ∈ [0,1] determines the brightness of the gain lines

Additional keyword arguments are sent to the function plotting the systems and can be used to specify colors, line styles etc. using regular Plots.jl syntax

This function is based on code subject to the two-clause BSD licence Copyright 2011 Will Robertson Copyright 2011 Philipp Allgeuer

ControlSystems.nicholsplot!

ControlSystems.ninputs

LinearAlgebra.norm — Function.

.. norm(sys, p=2; tol=1e-6)

norm(sys) or norm(sys,2) computes the H2 norm of the LTI system sys.

norm(sys, Inf) computes the L∞ norm of the LTI system sys. The H∞ norm is the same as the L∞ for stable systems, and Inf for unstable systems. If the peak gain frequency is required as well, use the function norminf instead.

tol is an optional keyword argument, used only for the computation of L∞ norms. It represents the desired relative accuracy for the computed L∞ norm (this is not an absolute certificate however).

sys is first converted to a state space model if needed.

The L∞ norm computation implements the 'two-step algorithm' in: N.A. Bruinsma and M. Steinbuch, 'A fast algorithm to compute the H∞-norm of a transfer function matrix', Systems and Control Letters 14 (1990), pp. 287-293. For the discrete-time version, see, e.g.,: P. Bongers, O. Bosgra, M. Steinbuch, 'L∞-norm calculation for generalized state space systems in continuous and discrete time', American Control Conference, 1991.

ControlSystems.normLinf_twoSteps_ct

ControlSystems.normLinf_twoSteps_dt

ControlSystems.norminf — Function.

.. (peakgain, peakgainfrequency) = norminf(sys; tol=1e-6)

Compute the L∞ norm of the LTI system sys, together with the frequency peakgainfrequency (in rad/TimeUnit) at which the gain achieves its peak value peakgain. The H∞ norm is the same as the L∞ for stable systems, and Inf for unstable systems.

tol is an optional keyword argument representing the desired relative accuracy for the computed L∞ norm (this is not an absolute certificate however).

sys is first converted to a state space model if needed.

The L∞ norm computation implements the 'two-step algorithm' in: N.A. Bruinsma and M. Steinbuch, 'A fast algorithm to compute the H∞-norm of a transfer function matrix', Systems and Control Letters 14 (1990), pp. 287-293. For the discrete-time version, see, e.g.,: P. Bongers, O. Bosgra, M. Steinbuch, 'L∞-norm calculation for generalized state space systems in continuous and discrete time', American Control Conference, 1991.

ControlSystems.noutputs

ControlSystems.nstates

ControlSystems.num

ControlSystems.numeric_type

ControlSystems.numpoly

ControlSystems.numvec

ControlSystems.nyquist — Function.

re, im, w = nyquist(sys[, w])

Compute the real and imaginary parts of the frequency response of system sys at frequencies w

re and im has size (length(w), ny, nu)

ControlSystems.nyquistplot — Function.

fig = nyquistplot(sys; gaincircles=true, kwargs...), nyquistplot(LTISystem[sys1, sys2...]; gaincircles=true, kwargs...)

Create a Nyquist plot of the LTISystem(s). A frequency vector w can be optionally provided.

gaincircles plots the circles corresponding to |S(iω)| = 1 and |T(iω)| = 1, where S and T are the sensitivity and complementary sensitivity functions.

kwargs is sent as argument to plot.

ControlSystems.nyquistplot!

ControlSystems.obsv — Function.

obsv(A, C) or obsv(sys)

Compute the observability matrix for the system described by (A, C) or sys.

Note that checking for observability by computing the rank from obsv is not the most numerically accurate way, a better method is checking if gram(sys, :o) is positive definite.

ControlSystems.parallel — Function.

series(sys1::LTISystem, sys2::LTISystem)

Connect systems in parallel, equivalent to sys2+sys1

ControlSystems.pid — Function.

Calculates and returns a PID controller on transfer function form. time indicates whether or not the parameters are given as gains (default) or as time constants series indicates whether or not the series form or parallel form (default) is desired

C = pid(; kp=0, ki=0; kd=0, time=false, series=false)

ControlSystems.pidplots — Function.

Plots interesting figures related to closing the loop around process P with a PID controller Send in a bunch of PID-parameters in any of the vectors kp, ki, kd. The vectors must be the same length.

time indicates whether or not the parameters are given as gains (default) or as time constants

series indicates whether or not the series form or parallel form (default) is desired

Available plots are :gof for Gang of four, :nyquist, :controller for a bode plot of the controller TF and :pz for pole-zero maps

One can also supply a frequency vector ω to be used in Bode and Nyquist plots

pidplots(P, args...; kps=0, kis=0, kds=0, time=false, series=false, ω=0)

See also loopshapingPI, stabregionPID

ControlSystems.place — Function.

place(A, B, p), place(sys::StateSpace, p)

Calculate gain matrix K such that the poles of (A-BK) in are in p

ControlSystems.plot

ControlSystems.pole — Function.

pole(sys)

Compute the poles of system sys.

ControlSystems.poly2vec

ControlSystems.poly_factors2string

ControlSystems.primitivetype

ControlSystems.print_compact

ControlSystems.print_siso

ControlSystems.printpoly2

ControlSystems.printpolydesc — Function.

Prints polynomial in descending order, with variable var

ControlSystems.printpolyfun

ControlSystems.pzmap — Function.

fig = pzmap(fig, system, args...; kwargs...)

Create a pole-zero map of the LTISystem(s) in figure fig, args and kwargs will be sent to the scatter plot command.

ControlSystems.pzmap!

ControlSystems.reduce_sys — Function.

Implements REDUCE in the Emami-Naeini & Van Dooren paper. Returns transformed A, B, C, D matrices. These are empty if there are no zeros.

ControlSystems.rlocus — Function.

rlocusplot(P::LTISystem, K)

Computes and plots the root locus of the SISO LTISystem P with a negative feedback loop and feedback gains K, if K is not provided, range(1e-6,stop=500,length=10000) is used. If OrdinaryDiffEq.jl is installed and loaded by the user (using OrdinaryDiffEq), rlocusplot will use an adaptive step-size algorithm to select values of K. A scalar Kmax can then be given as second argument.

ControlSystems.rlocusplot

ControlSystems.rlocusplot!

ControlSystems.roots2poly_factors

ControlSystems.roots2real_poly_factors

ControlSystems.rst

ControlSystems.rstc — Function.

See ?rstd for the discerte case

ControlSystems.rstd — Function.

rstd Polynomial synthesis in discrete time.

R,S,T=rstd(BPLUS,BMINUS,A,BM1,AM,AO,AR,AS)

R,S,T=rstd(BPLUS,BMINUS,A,BM1,AM,AO,AR)

R,S,T=rstd(BPLUS,BMINUS,A,BM1,AM,AO)

Polynomial synthesis according to CCS ch 10 to design a controller R(q) u(k) = T(q) r(k) - S(q) y(k)

Inputs: BPLUS : Part of open loop numerator BMINUS : Part of open loop numerator A : Open loop denominator BM1 : Additional zeros AM : Closed loop denominator AO : Observer polynomial AR : Pre-specified factor of R, e.g integral part [1, -1]^k AS : Pre-specified factor of S, e.g notch filter [1, 0, w^2]

Outputs: R,S,T : Polynomials in controller

See function DAB how the solution to the Diophantine- Aryabhatta-Bezout identity is chosen.

See Computer-Controlled Systems: Theory and Design, Third Edition Karl Johan Åström, Björn Wittenmark

ControlSystems.series — Function.

series(sys1::LTISystem, sys2::LTISystem)

Connect systems in series, equivalent to sys2*sys1

ControlSystems.setPlotScale — Function.

setPlotScale(str)

Set the default scale of magnitude in bodeplot and sigmaplot. str should be either "dB" or "log10".

ControlSystems.showterm2

ControlSystems.sigma — Function.

sv, w = sigma(sys[, w])

Compute the singular values of the frequency response of system sys at frequencies w

sv has size (length(w), max(ny, nu))

ControlSystems.sigmaplot — Function.

sigmaplot(sys, args...), sigmaplot(LTISystem[sys1, sys2...], args...)

Plot the singular values of the frequency response of the LTISystem(s). A frequency vector w can be optionally provided.

kwargs is sent as argument to Plots.plot.

ControlSystems.sigmaplot!

ControlSystems.siso_ss_to_zpk — Function.

Convert get zpk representation of sys from input j to output i

ControlSystems.siso_tf_to_ss

ControlSystems.sisomargin — Function.

ωgₘ, gₘ, ωϕₘ, ϕₘ = sisomargin{S<:Real}(sys::LTISystem, w::AbstractVector{S}; full=false, allMargins=false)

returns frequencies for gain margins, gain margins, frequencies for phase margins, phase margins

ControlSystems.sminreal — Function.

sminreal(sys)

Compute the structurally minimal realization of the state-space system sys. A structurally minimal realization is one where only states that can be determined to be uncontrollable and unobservable based on the location of 0s in sys are removed.

DiffEqBase.solve — Function.

sol = solve(s::AbstractSimulator, x0, tspan,  args...; kwargs...)

Simulate the system represented by s from initial state x0 over time span tspan = (t0,tf). args and kwargs are sent to the solve function from OrdinaryDiffEq, e.g., solve(s, x0, tspan, Tsit5(), reltol=1e-5) solves the problem with solver Tsit5() and relative tolerance 1e-5.

See also Simulator lsim

ControlSystems.ss — Function.

sys = ss(A, B, C, D, Ts=0) -> sys

Create a state-space model sys::StateSpace{T, MT<:AbstractMatrix{T}} where MT is the type of matrixes A,B,C,D and T the element type.

This is a continuous-time model if Ts is omitted or set to 0. Otherwise, this is a discrete-time model with sampling period Ts. Set Ts=-1 for a discrete-time model with unspecified sampling period.

sys = ss(D[, Ts, ...]) specifies a static gain matrix D.

ControlSystems.ss2tf

ControlSystems.ssdata

ControlSystems.stabregionPID — Function.

fig, kp, ki = stabregionPID(P, [ω]; kd=0, doplot = true)

Segments of the curve generated by this program is the boundary of the stability region for a process with transfer function P(s) The PID controller is assumed to be on the form kp +ki/s +kd s

The curve is found by analyzing P(s)*C(s) = -1 ⟹ |PC| = |P| |C| = 1 arg(P) + arg(C) = -π

If P is a function (e.g. s -> exp(-sqrt(s)) ), the stability of feedback loops using PI-controllers can be analyzed for processes with models with arbitrary analytic functions

See also stabregionPID, loopshapingPI, pidplots

Base.step — Function.

y, t, x = step(sys[, Tf]) or y, t, x = step(sys[, t])

Calculate the step response of system sys. If the final time Tf or time vector t is not provided, one is calculated based on the system pole locations.

y has size (length(t), ny, nu), x has size (length(t), nx, nu)

ControlSystems.stepplot — Function.

stepplot(sys[, Tf[,  Ts]])

Plot step response of sys with optional final time Tf and discretization time Ts. If not defined, suitable values are chosen based on sys.

ControlSystems.stepplot!

ControlSystems.struct_ctrb_obsv

ControlSystems.struct_ctrb_states

ControlSystems.tf — Function.

sys = tf(num, den, Ts=0), sys = tf(gain, Ts=0)

Create as a fraction of polynomials:

sys::TransferFunction{SisoRational{T,TR}} = numerator/denominator where T is the type of the coefficients in the polynomial.

num: the coefficients of the numerator polynomial. Either scalar or vector to create SISO systems or an array of vectors to create MIMO system.

den: the coefficients of the denominator polynomial. Either vector to create SISO systems or an array of vectors to create MIMO system.

Ts: Sample time or 0 for continuous system.

Other uses: tf(sys): Convert sys to tf form. tf("s"), tf("z"): Create the continous transferfunction s.

See also: zpk, ss

ControlSystems.to_matrix

ControlSystems.toeplitz

ControlSystems.tzero — Function.

tzero(sys)

Compute the invariant zeros of the system sys. If sys is a minimal realization, these are also the transmission zeros.

ControlSystems.unwrap

ControlSystems.unwrap!

ControlSystems.zpconv — Function.

zpc(a,r,b,s) form conv(a,r) + conv(b,s) where the lengths of the polynomials are equalized by zero-padding such that the addition can be carried out

ControlSystems.zpk — Function.

zpk(gain, Ts=0), zpk(num, den, k, Ts=0), zpk(sys)

Create transfer function on zero pole gain form. The numerator and denominator are represented by their poles and zeros.

sys::TransferFunction{SisoZpk{T,TR}} = k*numerator/denominator where T is the type of k and TR the type of the zeros/poles, usually Float64 and Complex{Float64}.

num: the roots of the numerator polynomial. Either scalar or vector to create SISO systems or an array of vectors to create MIMO system.

den: the roots of the denominator polynomial. Either vector to create SISO systems or an array of vectors to create MIMO system.

k: The gain of the system. Obs, this is not the same as dcgain.

Ts: Sample time or 0 for continuous system.

Other uses:

zpk(sys): Convert sys to zpk form.

zpk("s"): Create the transferfunction s.

ControlSystems.zpkdata — Function.

z, p, k = zpkdata(sys)

Compute the zeros, poles, and gains of system sys.

Returns

z : Matrix{Vector{ComplexF64}}, (ny x nu)

p : Matrix{Vector{ComplexF64}}, (ny x nu)

k : Matrix{Float64}, (ny x nu)