Docstrings

Word2Vec.Word2Vec

Word2Vec.WordClusters

Word2Vec.WordVectors

Word2Vec._from_binary

Word2Vec._from_text

Word2Vec.analogy — Function.

analogy(wv, pos, neg, n=5)

Compute the analogy similarity between two lists of words. The positions and the similarity values of the top n similar words will be returned. For example, king - man + woman = queen will be pos=["king", "woman"], neg=["man"].

Word2Vec.analogy_words — Function.

analogy_words(wv, pos, neg, n=5)

Return the top n words computed by analogy similarity between positive words pos and negaive words neg. from the WordVectors wv.

Word2Vec.clusters — Function.

clusters(wc)

Return all the clusters from the WordClusters wc.

Word2Vec.cosine — Function.

cosine(wv, word, n=10)

Return the position of n (by default n = 10) neighbors of word and their cosine similarities.

Word2Vec.cosine_similar_words — Function.

cosine_similar_words(wv, word, n=10)

Return the top n (by default n = 10) most similar words to word from the WordVectors wv.

Word2Vec.eval

Word2Vec.get_cluster — Function.

get_cluster(wc, word)

Return the cluster number for a word in the vocabulary.

Word2Vec.get_vector — Function.

get_vector(wv, word)

Return the vector representation of word from the WordVectors wv.

Word2Vec.get_words — Function.

get_words(wc, cluster)

For the WordCluster wc, return all the words from a given cluster number cluster.

Word2Vec.in_vocabulary — Function.

in_vocabulary(wv, word)

Return true if word is part of the vocabulary of the WordVector wv and false otherwise.

in_vocabulary(wc, word)

For the WordCluters wc, return true if word is part of the vocabulary of wc and false otherwise.

Word2Vec.include

Word2Vec.index — Function.

index(wv, word)

Return the index of word from the WordVectors wv.

index(wc, word)

Return the index of word from the WordCluaters wc.

Word2Vec.similarity — Function.

similarity(wv, word1, word2)

Return the cosine similarity value between two words word1 and word2.

Base.size — Function.

size(wv)

Return the word vector length and the number of words as a tuple.

Word2Vec.vocabulary — Function.

vocabulary(wv)

Return the vocabulary as a vector of words of the WordVectors wv.

vocabulary(wc)

Return all the vocabulary of the WordClusters wc.

Word2Vec.word2clusters — Function.

 word2cluster(train, output, classes; size=100, window=5, sample=1e-3, hs=0,  negative=5, threads=1, iter=5, min_count=5, alpha=0.025, debug=2, binary=1, cbow=1, save_vocal=Nothing(), read_vocab=Nothing(), verbose=false)

Parameters for training:
    train <file>
        Use text data from <file> to train the model
    output <file>
        Use <file> to save the resulting word vectors / word clusters
    size <Int>
        Set size of word vectors; default is 100
    window <Int>
        Set max skip length between words; default is 5
    sample <AbstractFloat>
        Set threshold for occurrence of words. Those that appear with
        higher frequency in the training data will be randomly
        down-sampled; default is 0 (off), useful value is 1e-5
    hs <Int>
        Use Hierarchical Softmax; default is 1 (0 = not used)
    negative <Int>
        Number of negative examples; default is 0, common values are 5 - 10
        (0 = not used)
    threads <Int>
        Use <Int> threads (default 1)
    iter <Int>
        Run more training iterations (default 5)
    min_count <Int>
        This will discard words that appear less than <Int> times
        (default 5)
    alpha <AbstractFloat>
        Set the starting learning rate; default is 0.025
    classes <Int>
        Number of word classes; if 0, output word classes rather than
        word vectors (default 0)
    debug <Int>
        Set the debug mode (default = 2 = more info during training)
    binary <Int>
        Save the resulting vectors in binary moded; default is 0 (off)
    cbow <Int>
        Use the continuous back of words model; default is 1 (0 for skip-gram
        model)
    save_vocab <file>
        The vocabulary will be saved to <file>
    read_vocab <file>
        The vocabulary will be read from <file>, not constructed from the
        training data
    verbose <Bool>
        Print output from training

Word2Vec.word2phrase — Function.

 word2phrase(train, output; min_count=5, threshold=100, debug=2)

Parameters for training:
train <file>
      Use text data from <file> to train the model 
output <file>
          Use <file> to save the resulting word vectors / 
          word clusters / phrases
min_count <Int>
          This will discard words that appear less than <int> times; 
          default is 5
threshold <AbstractFloat>
  	      The <AbstractFloat> value represents threshold for 
          forming the phrases (higher means less phrases); default 100
debug <Int>
      Set the debug mode (default = 2 = more info during training)

Word2Vec.word2vec — Function.

 word2vec(train, output; size=100, window=5, sample=1e-3, hs=0,  negative=5, threads=12, iter=5, min_count=5, alpha=0.025, debug=2, binary=1, cbow=1, save_vocal=Nothing(), read_vocab=Nothing(), verbose=false,)

Parameters for training:
    train <file>
        Use text data from <file> to train the model
    output <file>
        Use <file> to save the resulting word vectors / word clusters
    size <Int>
        Set size of word vectors; default is 100
    window <Int>
        Set max skip length between words; default is 5
    sample <AbstractFloat>
        Set threshold for occurrence of words. Those that appear with
        higher frequency in the training data will be randomly
        down-sampled; default is 1e-5.
    hs <Int>
        Use Hierarchical Softmax; default is 1 (0 = not used)
    negative <Int>
        Number of negative examples; default is 0, common values are 
        5 - 10 (0 = not used)
    threads <Int>
        Use <Int> threads (default 12)
    iter <Int>
        Run more training iterations (default 5)
    min_count <Int>
        This will discard words that appear less than <Int> times; default
        is 5
    alpha <AbstractFloat>
        Set the starting learning rate; default is 0.025
    debug <Int>
        Set the debug mode (default = 2 = more info during training)
    binary <Int>
        Save the resulting vectors in binary moded; default is 0 (off)
    cbow <Int>
        Use the continuous back of words model; default is 1 (skip-gram
        model)
    save_vocab <file>
        The vocabulary will be saved to <file>
    read_vocab <file>
        The vocabulary will be read from <file>, not constructed from the
        training data
    verbose <Bool>
        Print output from training

Word2Vec.wordclusters — Function.

wordclusters(fname)

Generate a WordClusters type object from the text file fname.

Word2Vec.wordvectors — Function.

wordvectors(filename [,type=Float64][; kind=:text, skip=false, normalize=true])

Generate a WordVectors type object from a file.

Arguments

filename::AbstractString the embeddings file name

type::Type type of the embedding vector elements; default Float64

Keyword arguments

kind::Symbol specifies whether the embeddings file is textual (:text)

or binary (:binary); default :text

skip::Bool in binary embeddings files specifies whether the newline

byte is missing or not (use true for Google pre-trained models); default false

normalize:Bool specifies whether to normalize the embedding vectors

i.e. return unit vectors; default true