Docstrings

Widgets.@auto — Macro.

@auto(expr)

Macro to automatize widget creation. Transforms x = rhs into x = widget(rhs, label = "x").

Widgets.@layout — Macro.

@layout(d, x)

Apply the expression x to the widget d, replacing e.g. symbol :s with the corresponding subwidget d[:s] In this context, _ refers to the whole widget. To use actual symbols, escape them with ^, as in ^(:a). @layout can be combined with @map to have the layout update interactively as function of some widget.

Examples

julia> using Interact

julia> cpt = OrderedDict(:vertical => Observable(true), :b => slider(1:100), :c => button());

julia> t = Widget{:test}(cpt, output = observe(cpt[:b]));

julia> Widgets.@layout t vbox(:b, CSSUtil.vskip(1em), :c);

julia> Widgets.@layout t Widgets.@map &(:vertical) ? vbox(:b, CSSUtil.vskip(1em), :c) : hbox(:b, CSSUtil.hskip(1em), :c);

Use @layout! to set the widget layout in place:

julia> @layout! t Widgets.@map &(:vertical) ? vbox(:b, CSSUtil.vskip(1em), :c) : hbox(:b, CSSUtil.hskip(1em), :c);

@layout(x)

Curried version of @layout(d, x): anonymous function mapping d to @layout(d, x).

Widgets.@layout! — Macro.

@layout!(d, x)

Set d.layout to match the result of Widgets.@layout(x). See Widgets.@layout for more information.

Examples

julia> using Interact

julia> t = Widget{:test}(OrderedDict(:b => slider(1:100), :c => button()));

julia> @layout! t hbox(:b, CSSUtil.hskip(1em), :c);

Widgets.@manipulate — Macro.

@manipulate expr

The @manipulate macro lets you play with any expression using widgets. expr needs to be a for loop. The for loop variable are converted to widgets using the widget function (ranges become slider, lists of options become togglebuttons, etc...). The for loop body is displayed beneath the widgets and automatically updated as soon as the widgets change value.

Use throttle = df to only update the output after a small time interval dt (useful if the update is costly as it prevents multiple updates when moving for example a slider).

Examples

using Colors

@manipulate for r = 0:.05:1, g = 0:.05:1, b = 0:.05:1
    HTML(string("<div style='color:#", hex(RGB(r,g,b)), "'>Color me</div>"))
end

@manipulate throttle = 0.1 for r = 0:.05:1, g = 0:.05:1, b = 0:.05:1
    HTML(string("<div style='color:#", hex(RGB(r,g,b)), "'>Color me</div>"))
end

@layout! can be used to adjust the layout of a manipulate block:

using Interact

ui = @manipulate throttle = 0.1 for r = 0:.05:1, g = 0:.05:1, b = 0:.05:1
    HTML(string("<div style='color:#", hex(RGB(r,g,b)), "'>Color me</div>"))
end
@layout! ui dom"div"(observe(_), vskip(2em), :r, :g, :b)
ui

Widgets.@nodeps — Macro.

@nodeps(expr)

Macro to remove need to depend on package X that defines a recipe to use it in one's own recipe. For example, InteractBase defines dropwdown recipe. To use dropdown in a recipe in a package, without depending on InteractBase, wrap the dropdown call in the @nodeps macro:

function myrecipe(i)
    label = "My recipe"
    wdg = Widgets.@nodeps dropdown(i)
    Widget(["label" => label, "dropdown" => wdg])
end

Widgets.AbstractBackend

Widgets.AbstractWidget

Widgets.DummyBackend

Widgets.Widget

Widgets.Widgets

Widgets._getindex

Widgets.accordion

Widgets.alert

Widgets.auto_helper!

Widgets.autocomplete

Widgets.backends

Widgets.button

Widgets.checkbox

Widgets.checkboxes

Widgets.colorpicker

Widgets.component

Widgets.components

Widgets.confirm

Widgets.datepicker

Widgets.defaultlayout

Widgets.div

Widgets.dropdown

Widgets.entry

Widgets.eval

Widgets.extract_name

Widgets.filepicker

Widgets.get_backend

Widgets.highlight

Widgets.include

Widgets.input

Widgets.isijulia

Widgets.isparameters

Widgets.isquotenode

Widgets.iswidgettuple

Widgets.latex

Widgets.layout — Function.

layout(f, w::Widget)

Create a new Widget that is a copy of w and whose layout is the layout of w composed with the function f.

Examples

using InteractBase, CSSUtil, Widgets
w = button("OK")
Widgets.layout(w) do t
    hbox("Click here", t)
end

layout(w::Widget)

Return the function that will be used to determine the layout of widget w.

Widgets.layout!

Widgets.layout!_helper

Widgets.layout_helper

Widgets.make_widget

Widgets.manipulatelayout

Widgets.map_block

Widgets.mask

Widgets.name2string

Widgets.node — Function.

node(w::Widget)

Return primary node for widget w

Widgets.notifications

Widgets.onchange — Function.

onchange(w::AbstractWidget, change = w[:changes])

Return a widget that's identical to w but only updates on change. For a slider it corresponds to releasing it and for a textbox it corresponds to losing focus.

Examples

sld = slider(1:100) |> onchange # update on release
txt = textbox("Write here") |> onchange # update on losing focuse

Widgets.parse_layout_call

Widgets.parse_layout_call!

Widgets.quotenode

Widgets.radiobuttons

Widgets.rangepicker

Widgets.rangeslider

Widgets.replace_wdg

Widgets.reset_backend!

Widgets.scope — Function.

scope(w::Widget)

Return primary scope for widget w if it exists, nothing otherwise.

Widgets.scope! — Function.

scope!(w::Widget, sc)

sets up a primary scope sc for widget w

Widgets.set_backend!

Widgets.slap_design!

Widgets.slider

Widgets.spinbox

Widgets.symbols

Widgets.tabs

Widgets.tabulator

Widgets.textarea

Widgets.textbox

Widgets.timepicker

Widgets.toggle

Widgets.togglebuttons

Widgets.togglecontent

Widgets.toggles

Widgets.tooltip!

Widgets.triggeredby

Widgets.triggeredby!

Widgets.wdglabel

Widgets.widget — Function.

widget(args...; kwargs...)

Automatically convert Julia types into appropriate widgets. kwargs are passed to the more specific widget function.

Examples

map(display, [
    widget(1:10),                 # Slider
    widget(false),                # Checkbox
    widget("text"),               # Textbox
    widget(1.1),                  # Spinbox
    widget([:on, :off]),          # Toggle Buttons
    widget(Dict("π" => float(π), "τ" => 2π)),
    widget(colorant"red"),        # Color picker
    widget(Dates.today()),        # Date picker
    widget(Dates.Time()),         # Time picker
    ]);

Widgets.widgettype