Docstrings

SymEngine.@funs

SymEngine.@init_constant

SymEngine.@vars — Macro.

Macro to define 1 or more variables in the main workspace.

Symbolic values are defined with _symbol. This is a convenience

Example

@vars x y z

SymEngine.AsciiArt — Type.

show symengine logo

SymEngine.Basic

SymEngine.BasicComplexNumber

SymEngine.BasicNumber

SymEngine.BasicOp

SymEngine.BasicRealNumber

SymEngine.BasicTrigFunction

SymEngine.BasicType

SymEngine.CDenseMatrix

SymEngine.CDenseMatrix_free

SymEngine.CMapBasicBasic

SymEngine.CMapBasicBasic_free

SymEngine.CSetBasic

SymEngine.CSetBasic_free

SymEngine.CVecBasic

SymEngine.CVecBasic_free

SymEngine.Catalan

SymEngine.E

SymEngine.EulerGamma

SymEngine.IM

SymEngine.IMPLEMENT_ONE_ARG_FUNC

SymEngine.IMPLEMENT_TWO_ARG_FUNC

SymEngine.N — Function.

Convert a SymEngine numeric value into a Julian number

SymEngine.NAN

SymEngine.PI

SymEngine.SymEngine

SymEngine.SymFunction

SymEngine.SymbolicType

SymEngine.__init__

SymEngine._convert

SymEngine._finalizer

SymEngine._flat

SymEngine._lambdify

SymEngine._symbol

SymEngine.as_numer_denom

SymEngine.ascii_art

SymEngine.basic_free

SymEngine.check_deps

SymEngine.complex_number_types

SymEngine.constant_map

SymEngine.cosd

SymEngine.cotd

SymEngine.cscd

SymEngine.cse

SymEngine.dense_matrix_FFLDU

SymEngine.dense_matrix_FFLU

SymEngine.dense_matrix_LDL

SymEngine.dense_matrix_LU

SymEngine.dense_matrix_LU_solve

SymEngine.dense_matrix_add_matrix

SymEngine.dense_matrix_add_scalar

SymEngine.dense_matrix_cols

SymEngine.dense_matrix_det

SymEngine.dense_matrix_eq

SymEngine.dense_matrix_eye

SymEngine.dense_matrix_get_basic

SymEngine.dense_matrix_inv

SymEngine.dense_matrix_mul_matrix

SymEngine.dense_matrix_mul_scalar

SymEngine.dense_matrix_ones

SymEngine.dense_matrix_rows

SymEngine.dense_matrix_rows_cols

SymEngine.dense_matrix_set_basic

SymEngine.dense_matrix_submatrix

SymEngine.dense_matrix_transpose

SymEngine.dense_matrix_zeros

SymEngine.deps_file

SymEngine.deps_in_file

SymEngine.done

SymEngine.eval

SymEngine.evalf

SymEngine.expand

SymEngine.fn_map

SymEngine.free_symbols — Function.

Return free symbols in an expression as a Set

SymEngine.get_args — Function.

Return arguments of a function call as a vector of Basic objects

SymEngine.get_class_from_id

SymEngine.get_error

SymEngine.get_libversion — Function.

Get libsymengine version

SymEngine.get_symbol — Function.

Helper function to lookup a symbol from libsymengine

SymEngine.get_symengine_class — Function.

Get SymEngine class of an object (e.g. 1=>:Integer, 1//2 =:Rational, sin(x) => :Sin, ...

SymEngine.get_type

SymEngine.have_component — Function.

Check whether libsymengine was compiled with comp

SymEngine.have_mpc

SymEngine.have_mpfr

SymEngine.include

SymEngine.init_constants

SymEngine.lambdify — Function.

lambdify

evaluates a symbolless expression or returns a function

SymEngine.lambertw

SymEngine.libgmp

SymEngine.libmpc

SymEngine.libmpfr

SymEngine.libsymengine

SymEngine.libversion

SymEngine.map_fn

SymEngine.next

SymEngine.nextprime

SymEngine.noinit

SymEngine.number_types

SymEngine.oo

SymEngine.op_types

SymEngine.prec

SymEngine.real_number_types

SymEngine.secd

SymEngine.series — Function.

Series expansion to order n about point x0

SymEngine.sind

SymEngine.start

SymEngine.subs — Function.

subs

Substitute values into a symbolic expression.

Examples

@vars x y
ex = x^2 + y^2
subs(ex, x, 1) # 1 + y^2
subs(ex, (x, 1)) # 1 + y^2
subs(ex, (x, 1), (y,x)) # 1 + x^2, values are substituted left to right.
subs(ex, x=>1)  # alternate to subs(x, (x,1))
subs(ex, x=>1, y=>1) # ditto

SymEngine.symbols — Function.

symbols(::Symbol) construct symbolic value

Examples:

a = symbols(:a)
x = symbols("x")
x,y = symbols("x y")
x,y,z = symbols("x,y,z")

SymEngine.tand

SymEngine.throw_if_error

SymEngine.toString

SymEngine.trig_types

SymEngine.walk_expression

SymEngine.zoo