Docstrings

DataValues.DataValue

DataValues.DataValueArray — Type.

DataValueArray{T, N} is an efficient alternative to Array{DataValue{T}, N}.

DataValues.DataValueException

DataValues.DataValueMatrix

DataValues.DataValueVector

DataValues.DataValuehash_seed

DataValues.DataValues

DataValues.NA

DataValues._default_splice

DataValues._nullable_eltype

DataValues._unsafe_get_eltype

DataValues.dropna — Function.

dropna(X::AbstractVector)

Return a vector containing only the non-missing entries of X, unwrapping DataValue entries. A copy is always returned, even when X does not contain any missing values.

DataValues.dropna! — Function.

dropna!(X::AbstractVector)

Remove missing entries of X in-place and return a Vector view of the unwrapped DataValue entries. If no missing values are present, this is a no-op and X is returned.

dropna!(X::DataValueVector)

Remove missing entries of X in-place and return a Vector view of the unwrapped DataValue entries.

DataValues.eval

DataValues.hasvalue

DataValues.include

DataValues.isna

DataValues.maptoTuple

DataValues.nullable_returntype

DataValues.padna — Function.

padna(X::DataValueVector, front::Integer, back::Integer)

return a copy of X with front null entries inserted at the beginning of the copy and back null entries inserted at the end.

DataValues.padna! — Function.

padna!(X::DataValueVector, front::Integer, back::Integer)

Insert front null entries at the beginning of X and add back null entries at the end of X. Returns X.

DataValues.unsafe_get