Docstrings

LambertW.BRANCH_POINT_COEFFS_FLOAT64

LambertW.LambertW

LambertW.__init__

LambertW._lambertw

LambertW._lambertw0

LambertW._lambertwm1

LambertW.branch_point_series

LambertW.cget

LambertW.compute_a_coeffs

LambertW.compute_branch_point_coeffs

LambertW.compute_m_coefficients

LambertW.cset

LambertW.eval

LambertW.finv — Function.

finv(::typeof(lambertw)) -> Function

return the functional inverse of the Lambert W function.

LambertW.horner

LambertW.include

LambertW.lambertw — Function.

lambertw(z::Complex{T}, k::V=0) where {T<:Real, V<:Integer}
lambertw(z::T, k::V=0) where {T<:Real, V<:Integer}

Compute the kth branch of the Lambert W function of z. If z is real, k must be either 0 or -1. For Real z, the domain of the branch k = -1 is [-1/e, 0] and the domain of the branch k = 0 is [-1/e, Inf]. For Complex z, and all k, the domain is the complex plane.

julia> lambertw(-1/e, -1)
-1.0

julia> lambertw(-1/e, 0)
-1.0

julia> lambertw(0, 0)
0.0

julia> lambertw(0, -1)
-Inf

julia> lambertw(Complex(-10.0, 3.0), 4)
-0.9274337508660128 + 26.37693445371142im

LambertW.lambertw_branch_one

LambertW.lambertw_branch_zero

LambertW.lambertw_root_finding

LambertW.lambertwbp — Function.

lambertwbp(z, k=0)

Compute accurate value of 1 + W(-1/e + z), for abs(z) in [0, 1/e] for k either 0 or -1. This function is faster and more accurate near the branch point -1/e between k=0 and k=1. The result is accurate to Float64 precision for abs(z) < 0.32. If k=-1 and imag(z) < 0, the value on the branch k=1 is returned.

julia> lambertw(-1/e + 1e-18, -1)
-1.0

julia> lambertwbp(1e-18, -1)
-2.331643983409312e-9

# Same result, but 1000 times slower
julia> convert(Float64, (lambertw(-BigFloat(1)/e + BigFloat(10)^(-18), -1) + 1))
-2.331643983409312e-9

Note

lambertwbp uses a series expansion about the branch point z=-1/e to avoid loss of precision. The loss of precision in lambertw is analogous to the loss of precision in computing the sqrt(1-x) for x close to 1.

LambertW.omega — Constant.

omega
ω

The constant defined by ω exp(ω) = 1.

julia> ω
ω = 0.5671432904097...

julia> omega
ω = 0.5671432904097...

julia> ω * exp(ω)
1.0

julia> big(omega)
5.67143290409783872999968662210355549753815787186512508135131079223045793086683e-01

LambertW.omega_const

LambertW.omega_const_

LambertW.omega_const_bf_

LambertW.wser100

LambertW.wser12

LambertW.wser19

LambertW.wser26

LambertW.wser290

LambertW.wser3

LambertW.wser32

LambertW.wser5

LambertW.wser50

LambertW.wser7

LambertW.ω — Constant.

omega
ω

The constant defined by ω exp(ω) = 1.

julia> ω
ω = 0.5671432904097...

julia> omega
ω = 0.5671432904097...

julia> ω * exp(ω)
1.0

julia> big(omega)
5.67143290409783872999968662210355549753815787186512508135131079223045793086683e-01