Docstrings

NNlib.@fix

NNlib.Dims

NNlib.MLN2

NNlib.NNlib

NNlib.__init__

NNlib.__inits__

NNlib.cdims

NNlib.col2im2d!

NNlib.col2im3d!

NNlib.col2im_2d!

NNlib.col2im_3d!

NNlib.conv

NNlib.conv!

NNlib.conv2d!

NNlib.conv2d_grad_w!

NNlib.conv2d_grad_x!

NNlib.conv3d!

NNlib.conv3d_grad_w!

NNlib.conv3d_grad_x!

NNlib.crosscor

NNlib.crosscor!

NNlib.dcdims

NNlib.depthwiseconv

NNlib.depthwiseconv!

NNlib.depthwiseconv2d!

NNlib.depthwiseconv2d_grad_w!

NNlib.depthwiseconv2d_grad_x!

NNlib.depthwisecrosscor

NNlib.depthwisecrosscor!

NNlib.dilation_dims

NNlib.elu — Function.

elu(x, α = 1) =
  x > 0 ? x : α * (exp(x) - 1)

Exponential Linear Unit activation function. See Fast and Accurate Deep Network Learning by Exponential Linear Units. You can also specify the coefficient explicitly, e.g. elu(x, 1).

NNlib.eval

NNlib.expand

NNlib.float2integer

NNlib.gemm!

NNlib.head

NNlib.ilogb2k

NNlib.im2col2d!

NNlib.im2col3d!

NNlib.im2col_2d!

NNlib.im2col_3d!

NNlib.im2col_dims

NNlib.include

NNlib.ldexp3k

NNlib.leakyrelu — Function.

leakyrelu(x) = max(0.01x, x)

Leaky Rectified Linear Unit activation function. You can also specify the coefficient explicitly, e.g. leakyrelu(x, 0.01).

NNlib.log_fast — Function.

log_fast(x)

Compute the natural logarithm of x. The inverse of the natural logarithm is the natural expoenential function exp(x)

NNlib.log_fast_kernel

NNlib.logsigmoid — Function.

logσ(x)

Return log(σ(x)) which is computed in a numerically stable way.

julia> logσ(0.)
-0.6931471805599453
julia> logσ.([-100, -10, 100.])
3-element Array{Float64,1}:
 -100.0
  -10.0
   -0.0

NNlib.logsoftmax — Function.

logsoftmax(xs) = log.(exp.(xs) ./ sum(exp.(xs)))

logsoftmax computes the log of softmax(xs) and it is more numerically stable than softmax function in computing the cross entropy loss.

NNlib.logsoftmax!

NNlib.logσ — Function.

logσ(x)

Return log(σ(x)) which is computed in a numerically stable way.

julia> logσ(0.)
-0.6931471805599453
julia> logσ.([-100, -10, 100.])
3-element Array{Float64,1}:
 -100.0
  -10.0
   -0.0

NNlib.max_pooling2d_bwd!

NNlib.max_pooling2d_fwd!

NNlib.max_pooling3d_bwd!

NNlib.max_pooling3d_fwd!

NNlib.maxpool

NNlib.maxpool!

NNlib.maxpool2d!

NNlib.maxpool2d_grad!

NNlib.maxpool3d!

NNlib.maxpool3d_grad!

NNlib.maxpool_cpu!

NNlib.mean_pooling2d_bwd!

NNlib.mean_pooling2d_fwd!

NNlib.mean_pooling3d_bwd!

NNlib.mean_pooling3d_fwd!

NNlib.meanpool

NNlib.meanpool!

NNlib.meanpool2d!

NNlib.meanpool2d_grad!

NNlib.meanpool3d!

NNlib.meanpool3d_grad!

NNlib.meanpool_cpu!

NNlib.padtuple

NNlib.pdims

NNlib.psize

NNlib.relu — Function.

relu(x) = max(0, x)

Rectified Linear Unit activation function.

NNlib.selu — Function.

selu(x) = λ * (x ≥ 0 ? x : α * (exp(x) - 1))

λ ≈ 1.0507
α ≈ 1.6733

Scaled exponential linear units. See Self-Normalizing Neural Networks.

NNlib.sigmoid — Function.

σ(x) = 1 / (1 + exp(-x))

Classic sigmoid activation function.

NNlib.softmax — Function.

softmax(xs) = exp.(xs) ./ sum(exp.(xs))

Softmax takes log-probabilities (any real vector) and returns a probability distribution that sums to 1.

If given a matrix it will treat it as a batch of vectors, with each column independent.

julia> softmax([1,2,3.])
3-element Array{Float64,1}:
  0.0900306
  0.244728
  0.665241

NNlib.softmax!

NNlib.softplus — Function.

softplus(x) = log(exp(x) + 1)

See Deep Sparse Rectifier Neural Networks.

NNlib.softsign — Function.

softsign(x) = x / (1 + |x|)

See Quadratic Polynomials Learn Better Image Features.

NNlib.swish — Function.

swish(x) = x * σ(x)

Self-gated actvation function. See Swish: a Self-Gated Activation Function.

NNlib.σ — Function.

σ(x) = 1 / (1 + exp(-x))

Classic sigmoid activation function.

NNlib.σ_stable

NNlib.∇conv_data

NNlib.∇conv_data!

NNlib.∇conv_filter

NNlib.∇conv_filter!

NNlib.∇depthwiseconv_data

NNlib.∇depthwiseconv_data!

NNlib.∇depthwiseconv_filter

NNlib.∇depthwiseconv_filter!

NNlib.∇logsoftmax

NNlib.∇maxpool

NNlib.∇maxpool!

NNlib.∇maxpool_cpu!

NNlib.∇meanpool

NNlib.∇meanpool!

NNlib.∇meanpool_cpu!

NNlib.∇softmax

NNlib.∇softmax!