Docstrings

NearestNeighbors.@NODE

NearestNeighbors.@POINT

NearestNeighbors.@POINT_UNCHECKED

NearestNeighbors.ArrayBuffers

NearestNeighbors.BallTree — Type.

BallTree(data [, metric = Euclidean(), leafsize = 10]) -> balltree

Creates a BallTree from the data using the given metric and leafsize.

NearestNeighbors.BruteTree — Type.

BruteTree(data [, metric = Euclidean()) -> brutetree

Creates a BruteTree from the data using the given metric.

NearestNeighbors.Chebyshev

NearestNeighbors.Cityblock

NearestNeighbors.DEBUG

NearestNeighbors.DataFreeTree — Type.

DataFreeTree(treetype, data[, reorderbufffer = similar(data), kwargs...]) -> datafreetree

Creates a DataFreeTree which wraps a KDTree or BallTree. Keywords arguments are passed to their respective constructors.

The KDTree or BallTree will be stored without a reference to the underlaying data. injectdata has to be used to re-link them to a data array before use.

NearestNeighbors.Euclidean

NearestNeighbors.Hamming

NearestNeighbors.HyperRectangle

NearestNeighbors.HyperSphere

NearestNeighbors.KDNode

NearestNeighbors.KDTree — Type.

KDTree(data [, metric = Euclidean(), leafsize = 10]) -> kdtree

Creates a KDTree from the data using the given metric and leafsize. The metric must be a MinkowskiMetric.

NearestNeighbors.Minkowski

NearestNeighbors.MinkowskiMetric

NearestNeighbors.NNTree

NearestNeighbors.NODES_VISITED

NearestNeighbors.NearestNeighbors

NearestNeighbors.NormMetric

NearestNeighbors.POINTS_VISITED

NearestNeighbors.POINTS_VISITED_UNCHECKED

NearestNeighbors.TreeData

NearestNeighbors._inrange

NearestNeighbors._knn

NearestNeighbors.add_node_visited

NearestNeighbors.add_point_visited

NearestNeighbors.add_point_visited_unchecked

NearestNeighbors.add_points_inrange!

NearestNeighbors.add_points_knn!

NearestNeighbors.addall

NearestNeighbors.always_false

NearestNeighbors.build_BallTree

NearestNeighbors.build_KDTree

NearestNeighbors.check_input

NearestNeighbors.check_k

NearestNeighbors.check_radius

NearestNeighbors.compute_bbox

NearestNeighbors.copy_svec

NearestNeighbors.create_bsphere

NearestNeighbors.encloses

NearestNeighbors.eval

NearestNeighbors.eval_diff

NearestNeighbors.eval_pow

NearestNeighbors.find_largest_spread

NearestNeighbors.find_split

NearestNeighbors.get_T

NearestNeighbors.get_leaf_range

NearestNeighbors.get_max_dim

NearestNeighbors.get_max_distance

NearestNeighbors.get_min_dim

NearestNeighbors.get_min_distance

NearestNeighbors.get_points_dim

NearestNeighbors.getleft

NearestNeighbors.getparent

NearestNeighbors.getright

NearestNeighbors.heap_sort_inplace!

NearestNeighbors.include

NearestNeighbors.injectdata — Function.

injectdata(datafreetree, data) -> tree

Returns the KDTree/BallTree wrapped by datafreetree, set up to use data for the points data.

NearestNeighbors.inrange — Function.

inrange(tree::NNTree, points, radius [, sortres=false]) -> indices

Find all the points in the tree which is closer than radius to points. If sortres = true the resulting indices are sorted.

NearestNeighbors.inrange_kernel!

NearestNeighbors.inrange_point!

NearestNeighbors.interpolate

NearestNeighbors.intersects

NearestNeighbors.isleaf

NearestNeighbors.knn — Function.

knn(tree::NNTree, points, k [, sortres=false]) -> indices, distances

Performs a lookup of the k nearest neigbours to the points from the data in the tree. If sortres = true the result is sorted such that the results are in the order of increasing distance to the point. skip is an optional predicate to determine if a point that would be returned should be skipped.

NearestNeighbors.knn_kernel!

NearestNeighbors.knn_point!

NearestNeighbors.n_ps

NearestNeighbors.percolate_down!

NearestNeighbors.point_index

NearestNeighbors.print_stats

NearestNeighbors.reorder_data!

NearestNeighbors.reset_stats

NearestNeighbors.select_spec!