Docstrings

FreeTypeAbstraction.FREE_FONT_LIBRARY

FreeTypeAbstraction.FTFont

FreeTypeAbstraction.FontExtent

FreeTypeAbstraction.FreeTypeAbstraction

FreeTypeAbstraction.Vec

FreeTypeAbstraction.__init__

FreeTypeAbstraction._font_paths

FreeTypeAbstraction.add_recursive

FreeTypeAbstraction.eval

FreeTypeAbstraction.family_name

FreeTypeAbstraction.findfont

FreeTypeAbstraction.fontpaths

FreeTypeAbstraction.freetype_can_read

FreeTypeAbstraction.freetype_extensions

FreeTypeAbstraction.ft_done

FreeTypeAbstraction.ft_init

FreeTypeAbstraction.getextent

FreeTypeAbstraction.glyphbitmap

FreeTypeAbstraction.include

FreeTypeAbstraction.kerning

FreeTypeAbstraction.loadchar

FreeTypeAbstraction.loaded_faces

FreeTypeAbstraction.match_font

FreeTypeAbstraction.newface

FreeTypeAbstraction.one_or_typemax

FreeTypeAbstraction.renderface

FreeTypeAbstraction.renderstring! — Function.

renderstring!(img::AbstractMatrix, str::String, face, pixelsize, y0, x0;
fcolor=one_or_typemax(T), bcolor=zero(T), halign=:hleft, valign=:vbaseline) -> Matrix

Render str into img using the font face of size pixelsize at coordinates y0,x0.

Arguments

y0,x0: origin is in upper left with positive y going down

fcolor: foreground color; typemax(T) for T<:Integer, otherwise one(T)

bcolor: background color; set to nothing for transparent

halign: :hleft, :hcenter, or :hright

valign: :vtop, :vcenter, :vbaseline, or :vbottom

FreeTypeAbstraction.setpixelsize

FreeTypeAbstraction.style_name

FreeTypeAbstraction.try_load

FreeTypeAbstraction.valid_fontpaths