Docstrings

TensorToolbox.==

TensorToolbox.MatrixCell — Type.

MatrixCell(N)

Cell of matrices of length N.

TensorToolbox.TTrank — Function.

TTrank(X::TTtensor,full=false)
TTrank(X::TensorCell,full=false)

Representational TT-rank of X. If cores of X are of size R{n-1} x In x Rn, returns (R1,...,R{n-1}) for full=false or (1,R1,....,Rn) for full=true.

TensorToolbox.TTsvd — Function.

TTsvd(X;tol=1e-8,reqrank=[])
TTsvd(X::TTtensor;tol=1e-8,reqrank=[])

Decomposes full tensor X into its TT format. If input is already a TTtensor, recompresses it.

TensorToolbox.TTtensor

TensorToolbox.TTtv — Function.

TTtv(X::TTtensor,u::VectorCell,mode::Int)
TTtv(X::TTtensor,U::Matrix,mode::Int)
TTtv(X::TensorCell,u::VectorCell)
TTtv(X::TensorCell,U::Matrix)

TT-tensor times vector. For N=ndims(X) and n=[mode,mode+1,...,N] contracts nth core of the TT-tensor with vector u[N-n+1]. If input is a matrix, then vectors u are its columns. Called on a subset of cores, it contracts all given cores to vectors.

TensorToolbox.TensorCell — Type.

TensorCell(N)

Cell of multidimensional arrays of length N.

TensorToolbox.TensorToolbox

TensorToolbox.VectorCell — Type.

VectorCell(N)

Cell of vectors of length N.

TensorToolbox.arrange — Function.

arrange(X[,n::Integer])
arrange(X[,P::Vector])

Arrange the rank-1 components of a ktensor: normalize the columns of the factor matrices and then sort the ktensor components by magnitude, greatest to least.

Arguments:

n: Absorb the weights into the nth factor matrix instead of lambda.

P: Rearrange the components of X according to the permutation P. P should be a permutation of 1 to ncomponents(X).

TensorToolbox.arrange! — Function.

arrange!(X[,n::Integer])
arrange!(X[,P::Vector])

Arrange the rank-1 components of a ktensor: normalize the columns of the factor matrices and then sort the ktensor components by magnitude, greatest to least. Rewrite ktensor.

Arguments:

n: Absorb the weights into the nth factor matrix instead of lambda.

P: Rearrange the components of X according to the permutation P. P should be a permutation of 1 to ncomponents(X).

TensorToolbox.blockdiag — Function.

blockdiag(X,Y)

Create block diagonal tensor where tensors X and Y are block elements. If X and Y are matrices, equal to blkdiag for sparse matrices.

Base.cat — Function.

cat(X,Y,n)

Concatente htensors X and Y by mode n.

TensorToolbox.check_vector_input — Function.

check_vector_input(input,dim,default_value)

Check whether input vector is of appropriate size or if input is number create vector out of it.

TensorToolbox.children — Function.

children(T[,node])

Matrix of children for each node of a dimtree T.

TensorToolbox.colspace — Function.

colspace(X; <keyword arguments>)

Column space basis.

Arguments:

X: Matrix.

method ∈ {"svd","lanczos","randsvd"} Method for SVD. Default: "svd".

maxrank::Integer: Maximal rank. Optional.

atol::Number: Drop singular values below atol. Default: 1e-8.

rtol::Number: Drop singular values below rtol*sigma_1. Optional.

p::Integer: Oversampling parameter used by lanczos and randsvd methods. Defaul p=10.

TensorToolbox.contract — Function.

contract(X,Y)
contract(X,indX,Y,indY[,perm])
contract(X::TensorCell)

Contracted product of tensors. Contract indX modes of array X to indY modes of array Y and permute the result by vector perm. Default: indX=[ndims(X)], indY=[1].

contract(T::TTtensor,X::Array,start_mode,modes)

Contract modes 1:ndims(X-1) of full tensor X to TTtensor T cores[startmode:startmode+modes-1]. Output: TTtensor.

TensorToolbox.coresize — Function.

coresize(X)

Size of core tensor of a ttensor.

TensorToolbox.count_leaves — Function.

count_leaves(T,node)

Number of leaves under a node of a dimtree T.

TensorToolbox.cp_als — Function.

cp_als(X,R;init,tol,maxit,dimorder)

Compute a CP decomposition with R components of a tensor X .

Arguments:

init ∈ {MatrixCell,"rand","nvecs","eigs"}. Initial guess for factor matrices. If init="nvecs" (same as "eigs") initialize matrices with function nvecs.

tol: Tolerance. Defualt: 1e-4.

maxit: Maximal number of iterations. Default: 1000.

`dimorder': Order of dimensions. Default: 1:ndims(A).

TensorToolbox.diagt — Function.

diagt(v[,dims])

Create a diagonal tensor for a given vector of diagonal elements. Generalization of diagm.

TensorToolbox.dims — Function.

dims(T,node)

Content of a node of a dimtree T.

TensorToolbox.dimtree — Type.

dimtree(leaves::Vector{Int}[,internal_nodes::Vector{Int}])
dimtree(N[,treetype])

Dimension tree. Create from:

a vector of leaves (and a vector of internal nodes),

an order of a tensor N and a type treetype of a dimtree. Default: treetype="balanced".

TensorToolbox.dirsum — Function.

dirsum(X,Y)

Direct sum od tensors. Equal to blockdiag.

Base.Multimedia.display — Function.

TensorToolbox:

display(X::ttensor[,name])
display(X::ktensor[,name])
display(X::htensor[,name])

Displays a tensor X of a given name.

display(T)

Display a dimtree T.

Base.dropdims — Function.

TensorToolbox:

dropdims(X)

Remove singleton dimensions from htensor.

TensorToolbox.eval

TensorToolbox.ewprod — Function.

ewprod(X,Y)

Element-wise product of two ttensors.

TensorToolbox.extract — Function.

extract(X,factors)

Create a new ktensor with only the specified factors.

TensorToolbox.eye — Function.

eye(n::Integer)

Identity matrix of size nxn.

TensorToolbox.fixsigns — Function.

fixsign(X)

Fix sign ambiguity of a ktensor.

TensorToolbox.fixsigns! — Function.

fixsign(X)

Fix sign ambiguity of a ktensor. Rewrite ktensor.

TensorToolbox.full — Function.

TensorToolbox:

full(X::ttensor)
full(X::ktensor)
full(X::htensor)

Make full tensor out of a decomposed tensor.

TensorToolbox.hadcten — Function.

hadcten(X,Y,fmat)

Core tensor of Hadamard product of two ttensors with given factor matrices.

TensorToolbox.height — Function.

height(T)

Height of a dimtree T.

TensorToolbox.hosvd — Function.

hosvd(X; <keyword arguments>)

Higher-order singular value decomposition.

Arguments:

X: Tensor (multidimensional array) or ttensor.

method ∈ {"svd","lanczos","randsvd"} Method for SVD. Default: "svd".

reqrank::Vector: Requested mutlilinear rank. Optional.

eps_abs::Number/Vector: Drop singular values (of mode-n matricization) below eps_abs. Default: 1e-8.

eps_rel::Number/Vector: Drop singular values (of mode-n matricization) below epsrel*sigma1. Optional.

p::Integer: Oversampling parameter. Defaul p=10.

TensorToolbox.hosvd1 — Function.

hosvd1(X,Y; <keyword arguments>)

Hadamard product of ttensors X and Y as ttensor. Creates the product and calls hosvd. See also: hosvd2, hosvd3, hosvd4.

Arguments:

method ∈ {"svd","lanczos","randsvd"}. Method for SVD. Default: "randsvd".

reqrank::Vector: Requested mutlilinear rank. Optional.

eps_abs::Number/Vector: Drop singular values (of mode-n matricization) below eps_abs. Optional.

eps_rel::Number/Vector: Drop singular values (of mode-n matricization) below epsrel*sigma1. Optional.

p::Integer: Oversampling parameter. Defaul p=10.

TensorToolbox.hosvd2 — Function.

hosvd2(X,Y; <keyword arguments>)

Hadamard product of ttensors X and Y as ttensor. Orthogonalizes factor matrices from structure and calls hosvd on updated core tensor. See also: hosvd1, hosvd3, hosvd4.

Arguments:

method ∈ {"svd","lanczos","randsvd"} Method for SVD. Default: "randsvd".

reqrank::Vector: Requested mutlilinear rank. Optional.

eps_abs::Number/Vector: Drop singular values (of mode-n matricization) below eps_abs. Optional.

eps_rel::Number/Vector: Drop singular values (of mode-n matricization) below epsrel*sigma1. Optional.

p::Integer: Oversampling parameter. Defaul p=10.

TensorToolbox.hosvd3 — Function.

hosvd3(X,Y; <keyword arguments>)

Hadamard product of ttensors X and Y as ttensor. Structure exploiting, works with (X ∗ Y)ₙ(X ∗ Y)ₙᵀ matrices. See also: hosvd1, hosvd2, hosvd4.

Arguments:

method ∈ {"lanczos","randsvd"} Structure exploiting method for SVD. Default: "lanczos".

reqrank::Vector: Requested mutlilinear rank. Optional.

variant ∈ {'A','B'} Variant of multiplication (X ∗ Y)ₙ(X ∗ Y)ₙᵀ. Default: 'B'.

eps_abs::Number/Vector: Drop singular values (of mode-n matricization) below eps_abs. Optional.

eps_rel::Number/Vector: Drop singular values (of mode-n matricization) below epsrel*sigma1. Optional.

p::Integer: Oversampling parameter. Defaul p=10.

TensorToolbox.hosvd4 — Function.

hosvd4(X,Y; <keyword arguments>)

Hadamard product of ttensors X and Y as ttensor. Uses rank-1 randomized algorithm for finding range of (X ∗ Y)ₙ. If reqrank defined, calls additonal hosvd on updated core tensor. See also: hosvd1, hosvd2, hosvd3.

Arguments:

method ∈ {"svd","lanczos","randsvd"} Method for SVD. Default: "svd".

reqrank::Vector: Requested mutlilinear rank. Optional.

eps_abs::Number/Vector: Drop singular values (of mode-n matricization) below eps_abs. Optional.

eps_rel::Number/Vector: Drop singular values (of mode-n matricization) below epsrel*sigma1. Optional.

p::Integer: Oversampling parameter. Defaul p=10.

TensorToolbox.hrank — Function.

fmat(X)

Hierarchical ranks of a htensor.

TensorToolbox.htensor — Type.

htensor(tree,trten,fmat)
htensor(tree,trten,U1,U2,...)

Hierarchical Tucker tensor.

Arguments:

tree::dimtree: Dimension tree.

trten::TensorCell: Transfer tensors. One for each internal node of the tree.

fmat::MatrixCell: Frame matrices. One for each leaf of the tree.

For htensor X, X.isorth=true if frame matrices are othonormal.

TensorToolbox.htrunc — Function.

htrunc(X[,tree])

Truncate full tensor X into a htensor for a given tree. If tree not specified use balanced tree.

TensorToolbox.include

TensorToolbox.innerprod — Function.

innerprod(X,Y) innerprod(X::ttensor,Y::ttensor) innerprod(X::ktensor,Y::ktensor)

Inner product of two tensors.

TensorToolbox.is_leaf — Function.

is_leaf(T,node)

True if a node is a leaf in a dimtree T. False otherwise.

TensorToolbox.is_left — Function.

is_left(T,node)

True if a node is a left node of dimtree T. False otherwise.

TensorToolbox.is_right — Function.

is_right(T,node)

True if a node is a right node of dimtree T. False otherwise.

Base.isequal — Function.

isequal(X::ttensor,Y::ttensor)
isequal(X::ktensor,Y::ktensor)
isequal(X::ktensor,Y::htensor)

Two tensors in decomposed format are equal if they have equal components. Same as: X==Y.

isequal(T1,T2)

True if two dimensional trees are equal, false otherwise.

TensorToolbox.khatrirao — Function.

khatrirao(M,t='n')
khatrirao(M,n,t='n')

Khatri-Rao product of matrices from M: M₁⊙ ⋯ ⊙ Mₙ. Optionally skip nth matrix. If t='t', compute transpose Khatri-Rao product: M₁⊙ᵀ ⋯ ⊙ᵀ Mₙ.

TensorToolbox.krank — Function.

krank(X::ktensor)

Represenation k-rank. Equal to number of columns of factor matrices if ktensor X.

TensorToolbox.kron

TensorToolbox.krontkron — Function.

krontkron(A,v,t='n')

Kronecker product of matrices multiplied by Kronecker product of vectors.

TensorToolbox.krontm — Function.

krontm(X,Y,M[,modes,t='n'])

Kronecker product of two tensors times matrix (n-mode product): (X ⊗ Y) x₁ M₁ x₂ M₂ x₃ ⋯ xₙ Mₙ.

Arguments:

X::Array

Y::Array

M::Matrix/MatrixCell

modes::Integer/Vector : Modes for multiplication. Default: 1:length(M).

t='t': Transpose matrices from M.

TensorToolbox.krontv — Function.

krontv(A,B,v)

Kronecker product times vector: (A ⊗ B)v. If v is a matrix, multiply column by column.

TensorToolbox.krtv — Function.

krtv(A,B,v)

Khatri-Rao product times vector: (A ⊙ B)v. If v is a matrix, multiply column by column.

TensorToolbox.kten2TT — Function.

kten2TT(X::ktensor)

Transform ktensor to TTtensor.

TensorToolbox.ktensor — Type.

ktensor(fmat)
ktensor(lambda,fmat)

Tensor in Kruskal format defined by its vector of weights lambda and factor matrices. Default lambda: vector of ones. For ktensor X, X.isorth=true if factor matrices are othonormal.

TensorToolbox.lanczos — Function.

lanczos(A; <keyword arguments>)

Lanczos based SVD - computes left singular vectors and singular values of a matrix.

Arguments:

A::Matrix

tol: Tolerance - discard singular values below tol. Default: 1e-8.

maxit: Maximal number of iterations. Default: 1000.

reqrank: Number of singular values and singular vectors to compute. Optional.

p: Oversampling parameter. Default: p=10.

lanczos(X,Y,n; <keyword arguments>)

Structure exploiting Lanczos based SVD - computes left singular vectors and singular values of n-mode matricization (X ∗ Y)ₙ. Works with matrix (X ∗ Y)ₙ(X ∗ Y)ₙᵀ.

Arguments:

reqrank::Integer: Requested rank. Optional.

variant ∈ {'A','B'} Variant of multiplication (X ∗ Y)ₙ(X ∗ Y)ₙᵀ. Default: 'B'.

tol::Number/Vector: Tolerance. Default: 1e-8.

maxit: Maximal number of iterations. Default: 1000.

p::Integer: Oversampling parameter. Defaul p=10.

TensorToolbox.lanczos_tridiag — Function.

lanczos_tridiag(A; <keyword arguments>)

Lanczos tridiagonalization algorithm - returns orthonormal Q and symmetric tridiagonal T such that A≈QTQ'.

Arguments:

A::Matrix

tol: Tolerance. Default: 1e-8.

maxit: Maximal number of iterations. Default: 1000.

reqrank: Number of singular values and singular vectors to compute. Optional.

p: Oversampling parameter. Default: p=10.

lanczos_tridiag(X,Y,n; <keyword arguments>)

Structure exploiting Lanczos tridiagonalization algorithm for n-mode matricization A=(X ∗ Y)ₙ(X ∗ Y)ₙᵀ. Returns orthonormal Q and symmetric tridiagonal T such that A≈QTQ'.

Arguments:

reqrank::Integer: Requested rank. Optional.

variant ∈ {'A','B'} Variant of multiplication (X ∗ Y)ₙ(X ∗ Y)ₙᵀ. Default: 'B'.

tol::Number/Vector: Tolerance. Default: 1e-8.

maxit: Maximal number of iterations. Default: 1000.

p::Integer: Oversampling parameter. Defaul p=10.

TensorToolbox.left_child_length — Function.

left_child_length(T[,node])

Length of the left child of a node of a dimtree T. Default: node=1.

TensorToolbox.lvl — Function.

lvl(T[,node])

Vector of levels for each node of a dimtree T.

TensorToolbox.matten — Function.

matten(A,n,dims)
matten(A,R,C,dims)

Fold matrix A into a tensor of dimension dims by mode n or by row and column vectors R and C.

TensorToolbox.mhadtm

TensorToolbox.mhadtv — Function.

mhadtv(X,Y,v,n,t='b';variant='B')

Mode-n matricized Hadamard product of ttensors X and Y times vector v.

t='b': (X ∗ Y)ₙ(X ∗ Y)ₙᵀv. Variant of multiplication (X ∗ Y)ₙ(X ∗ Y)ₙᵀ can be specified to 'A' or 'B'.

t='n': (X ∗ Y)ₙᵀv.

t='t': (X ∗ Y)ₙᵀv.

If v is a matrix, multiply column by column.

TensorToolbox.minus — Function.

minus(X::ttensor,Y::ttensor)
minus(X::ktensor,Y::ktensor)
minus(X::htensor,Y::htensor)

Subtraction of two tensors. Same as: X-Y.

minus(X::TTtensor,Y::TTtensor)

Subtraction of two TTtensors. Same as: X-Y.

TensorToolbox.mkrontm

TensorToolbox.mkrontv — Function.

mkrontv(X,Y,v,n,t='n')

Matricized Kronecker product of tensors X and Y times vector v (n-mode multiplication): (X ⊗ Y)ₙv. If t='t', transpose matricized Kronecker product: (X ⊗ Y)ᵀₙv. If v is a matrix, multiply column by column.

TensorToolbox.mrank — Function.

mrank(X[,tol])
mrank(X::ttensor[,tol])

Multilinear rank of a tensor with optionally given tolerance.

TensorToolbox.msvdvals — Function.

msvdvals(X,n)

Singular values of mode-n matricization of a ttensor calculated directly. See also: nvecs.

TensorToolbox.mtimes — Function.

mtimes(a,X)

Scalar times ttensor. Same as: a*X.

mtimes(a,X)

Scalar times TTtensor. Same as: a*X.

TensorToolbox.mtimes!

TensorToolbox.mttkrp — Function.

mttkrp(X,M,n)
mttkrp(X::ttensor,M,n)
mttkrp(X::ktensor,M,n)

Mode-n matricized tensor X times Khatri-Rao product of matrices from M (except nth) in reverse order.

TensorToolbox.ncomponents — Function.

ncomponents(X)

Number of components of a ktensor.

TensorToolbox.ndims

TensorToolbox.next_single_node

TensorToolbox.neye — Function.

neye(dims)

Identity tensor of a given dimension. Generalization of eye.

TensorToolbox.node2ind — Function.

node2ind(T,nodes)

Convert node numbers to transfer tensor or frames indices in a dimtree T.

TensorToolbox.nodes_on_lvl — Function.

nodes_on_lvl(T,l)

Nodes on a level l in a dimtree T.

TensorToolbox.non — Function.

non(T)

Number of nodes of a dimtree T.

TensorToolbox.norm

LinearAlgebra.normalize — Function.

normalize(X[,n;normtype,factor])

Normalize columns of factor matrices of a ktensor. Also ensures that lambda is positive.

Arguments:

n: Absorbe the excess weight into nth factor matrix. If n=0, equally divide the weights across the factor matrices and set lambda to vector of ones. If n="sort", sort the components by lambda value, from greatest to least.

normtype: Default: 2.

factor: Just normalize specified factor.

LinearAlgebra.normalize! — Function.

normalize!(X[,n,normtype,factor])

Normalize columns of factor matrices of a ktensor. Rewrite ktensor. Also ensures that lambda is positive.

Arguments:

n: Absorbe the excess weight into nth factor matrix. If n=0, equally divide the weights across the factor matrices and set lambda to vector of ones. If n="sort", sort the components by lambda value, from greatest to least.

normtype: Default: 2.

factor: Just normalize specified factor.

TensorToolbox.nrank — Function.

nrank(X,n[,tol])
nrank(X::ttensor,n[,tol])

Rank of the n-mode matricization of a tensor X (n-rank).

TensorToolbox.nvecs — Function.

nvecs(X,n,r=0;flipsign=false,svds=false)
nvecs(X::ttensor,n,r=0;flipsign=false)
nvecs(X::ktensor,n,r=0;flipsign=false)

Computes the r leading singular vectors of mode-n matricization of a tensor X. Works with XₙXₙᵀ.

Arguments:

flipsign=true: Make the largest magnitude element be positive.

svds=true: Use svds on Xₙ rather than eigs on XₙXₙᵀ.

Base.parent — Function.

parent(T[,node])

Vectors of parents for each node of a dimtree T.

TensorToolbox.permutedims

TensorToolbox.plus — Function.

plus(X::ttensor,Y::ttensor)
plus(X::ktensor,Y::ktensor)
plus(X::htensor,Y::htensor)

Addition of two tensors. Same as: X+Y.

plus(X::TTtensor,Y::TTtensor)

Addition of two TTtensors. Same as: X+Y.

TensorToolbox.positions — Function.

positions(T)

Positions of nodes of a dimtree T in a full binary tree.

TensorToolbox.randTTtensor — Function.

randTTtensor(I::Vector,R::Vector)
randTTtensor(I::Integer,R::Integer,N::Integer)

Create random TTtensor of size I and TT-rank rank R, or of order N and size I × ⋯ × I and TT-rank (R,...,R).

TensorToolbox.randhtensor — Function.

randhtensor(I::Vector,R::Vector)
randhtensor(I::Vector,R::Vector,T::dimtree)

Create random htensor of size I, hierarchical rank R and balanced tree T. If T not specified use balanced tree.

TensorToolbox.randktensor — Function.

randktensor(I::Vector,R::Integer)
randktensor(I::Integer,R::Integer,N::Integer)

Create random Kruskal tensor of size I with R components, or of order N and size I × ⋯ × I.

TensorToolbox.randrange — Function.

randrange(X,Y,n; <keyword arguments>)

Structure exploiting randomized range approximation of n-mode matricization of Hadamard product (X ∗ Y)ₙ, where X and Y are ttensors.

Arguments:

reqrank::Integer: Requested rank. Optional.

tol::Number/Vector: Tolerance. Default: 1e-8.

maxit: Maximal number of iterations. Default: 1000.

r: Number of samples for stopping criterion. Default: r=10.

p::Integer: Oversampling parameter. Defaul p=10.

TensorToolbox.randsvd — Function.

randsvd(A; <keyword arguments>)

Randomized SVD algorithm - returns left singular vectors and singular values of a matrix.

Arguments:

A::Matrix

tol: Tolerance - discard singular values below tol. Default: 1e-8.

maxit: Maximal number of iterations. Default: 1000.

reqrank: Number of singular values and singular vectors to compute. Optional.

r: Number of samples for stopping criterion. Default: r=10.

p: Oversampling parameter. Default: p=10.

randsvd(X,Y,n; <keyword arguments>)

Structure exploiting randomized SVD - computes left singular vectors and singular values of n-mode matricization (X ∗ Y)ₙ. Works with matrix (X ∗ Y)ₙ(X ∗ Y)ₙᵀ.

Arguments:

reqrank::Integer: Requested rank. Optional.

variant ∈ {'A','B'} Variant of multiplication (X ∗ Y)ₙ(X ∗ Y)ₙᵀ. Default: 'B'.

tol::Number/Vector: Tolerance. Default: 1e-8.

maxit: Maximal number of iterations. Default: 1000.

r: Number of samples for stopping criterion. Default: r=10.

p::Integer: Oversampling parameter. Defaul p=10.

TensorToolbox.randttensor — Function.

randttensor(I::Vector,R::Vector)
randttensor(I::Integer,R::Integer,N::Integer)

Create random ttensor of size I and multilinear rank R, or of order N and size I × ⋯ × I and mulilinear rank (R,...,R).

TensorToolbox.redistribute — Function.

redistribute(X,n)

Distribute lambda values to a mode n.

TensorToolbox.redistribute! — Function.

redistribute!(X,n)

Distribute lambda values to a mode n. Rewrite ktensor X.

TensorToolbox.reorth — Function.

  reorth(X::ttensor)
  reorth(X::htensor)
  reorth(X::TTtensor[,direction,full])

Orthogonalize factor matrices of a tensor.

TensorToolbox.reorth! — Function.

reorth!(X::ttensor)
reorth!(X::htensor)
reorth(X::TTtensor[,direction,full])

Orthogonalize factor matrices of a tensor. Rewrite ttensor.

TensorToolbox.show

TensorToolbox.sibling — Function.

sibling(T,node)

Sibling of a node in a dimtree T.

TensorToolbox.size

TensorToolbox.sthosvd — Function.

sthosvd(X,reqrank,p)

Sequentially truncated HOSVD of a tensor X of predifined rank and processing order p.

TensorToolbox.structure — Function.

structure(T)

For each internal node of a given dimtree, returns its representation t, its left child and right child representations tl and tr.

TensorToolbox.subnodes — Function.

subnodes(T,node)

Nodes in the subtree of a node in a dimtree T.

TensorToolbox.subtree — Function.

subtree(T,node)

Dimensional tree with the same structure as a subtree of a node in a dimtree T. See also: subnodes.

TensorToolbox.tenmat — Function.

tenmat(X,n)
tenmat(X,row=[],col=[])
tenmat(X::ttensor,n)
tenmat(X::ktensor,n)

Mode-n matricization of a tensor or matricization by row and column vectors R and C.

TensorToolbox.tkron — Function.

tkron(X,Y)

Kronecker product of two tensors X and Y. Direct generalization of Kronecker product of matrices.

TensorToolbox.tkrtv — Function.

tkrtv(A,B,v)

Transpose Khatri-Rao product times vector: (A ⊙ᵀ B)v. If v is a matrix, multiply column by column.

TensorToolbox.tocell — Function.

tocell(X[,n])

Converts ktensor X into MatrixCell. If n specified, absorb the weights into the nth factor matrix.

TensorToolbox.trten2mat — Function.

trten2mat(B::Array)
trten2mat(B::TensorCell)

Transfer tensor to matrix. If transfer tensor is given as a tensor of order 3 and size (r1,r2,r3), reshape it into a matrix of size (r1r2,r3).

TensorToolbox.trten2ten — Function.

trten2ten(B::Matrix,r1::Integer,r2::Integer)
trten2ten(B::MatrixCell,r1::Vector,r2::Vector)

Transfer tensor to tensor. If transfer tensor is given as a matrix, reshape it into a tensor of order 3 and size r1×r2×r3, where r3=size(B,2).

TensorToolbox.ttensor — Type.

ttensor(cten,fmat)

Tensor in Tucker format defined by its core tensor and factor matrices. For ttensor X, X.isorth=true if factor matrices are othonormal.

TensorToolbox.ttm — Function.

ttm(X,M[,modes,t='n'])
ttm(X::ttensor,M[,modes,t='n'])

Tensor times matrix (n-mode product): X x₁ M₁ x₂ M₂ x₃ ⋯ xₙ Mₙ Default modes: 1:length(M). If t='t', transpose matrices from M.

TensorToolbox.ttt — Function.

ttv(X,Y)

Outer product of two tensors.

TensorToolbox.ttv — Function.

ttv(X,V[,modes])

Tensor times vectors (n-mode product): X x₁ V₁ x₂ V₂ x₃ ⋯ xₙ Vₙ. Default modes: 1:length(M).

TensorToolbox.uminus — Function.

uminus(X::ttensor) uminus(X::ktensor) uminus(X::htensor)

Unary minus. Same as: (-1)*X.