Docstrings

BenchmarkTools.@belapsed — Macro.

@belapsed expression [other parameters...]

Similar to the @elapsed macro included with Julia, this returns the elapsed time (in seconds) to execute a given expression. It uses the @benchmark macro, however, and accepts all of the same additional parameters as @benchmark. The returned time is the minimum elapsed time measured during the benchmark.

BenchmarkTools.@benchmark

BenchmarkTools.@benchmarkable

BenchmarkTools.@btime — Macro.

@btime expression [other parameters...]

Similar to the @time macro included with Julia, this executes an expression, printing the time it took to execute and the memory allocated before returning the value of the expression.

Unlike @time, it uses the @benchmark macro, and accepts all of the same additional parameters as @benchmark. The printed time is the minimum elapsed time measured during the benchmark.

BenchmarkTools.@tagged

BenchmarkTools.BENCHMARKTOOLS_VERSION

BenchmarkTools.Benchmark

BenchmarkTools.BenchmarkGroup

BenchmarkTools.BenchmarkTools

BenchmarkTools.DEFAULT_PARAMETERS

BenchmarkTools.EVALS

BenchmarkTools.Parameters

BenchmarkTools.RESOLUTION

BenchmarkTools.SUPPORTED_TYPES

BenchmarkTools.TagFilter

BenchmarkTools.Trial

BenchmarkTools.TrialEstimate

BenchmarkTools.TrialJudgement

BenchmarkTools.TrialRatio

BenchmarkTools.VERSIONS

BenchmarkTools._lineartrial

BenchmarkTools._run

BenchmarkTools._show

BenchmarkTools._summary

BenchmarkTools.addgroup!

BenchmarkTools.allocs

BenchmarkTools.andexpr

BenchmarkTools.andreduce

BenchmarkTools.badext

BenchmarkTools.benchmarkable_parts

BenchmarkTools.collectvars

BenchmarkTools.colormap

BenchmarkTools.createchild!

BenchmarkTools.estimate_overhead

BenchmarkTools.eval

BenchmarkTools.filtervals

BenchmarkTools.filtervals!

BenchmarkTools.gcratio

BenchmarkTools.gcscrub

BenchmarkTools.gctime

BenchmarkTools.generate_benchmark_definition

BenchmarkTools.guessevals

BenchmarkTools.hasevals

BenchmarkTools.improvements

BenchmarkTools.include

BenchmarkTools.invariants

BenchmarkTools.isimprovement

BenchmarkTools.isinvariant

BenchmarkTools.isregression

BenchmarkTools.judge

BenchmarkTools.keyunion

BenchmarkTools.leaves

BenchmarkTools.leaves!

BenchmarkTools.lineartrial

BenchmarkTools.load

BenchmarkTools.loadparams!

BenchmarkTools.loadtagged!

BenchmarkTools.logistic

BenchmarkTools.mapvals

BenchmarkTools.mapvals!

BenchmarkTools.memory

BenchmarkTools.nullfunc

BenchmarkTools.overhead_sample

BenchmarkTools.params

BenchmarkTools.prettydiff

BenchmarkTools.prettymemory

BenchmarkTools.prettypercent

BenchmarkTools.prettytime

BenchmarkTools.printmemoryjudge

BenchmarkTools.printtimejudge

BenchmarkTools.prunekwargs

BenchmarkTools.quasiquote!

BenchmarkTools.ratio

BenchmarkTools.recover

BenchmarkTools.regressions

BenchmarkTools.rmskew

BenchmarkTools.rmskew!

BenchmarkTools.run_result

BenchmarkTools.sample

BenchmarkTools.save

BenchmarkTools.skewcutoff

BenchmarkTools.tagpredicate!

BenchmarkTools.tagrepr

BenchmarkTools.tagunion

BenchmarkTools.trim

BenchmarkTools.tune!

BenchmarkTools.warmup

BenchmarkTools.withtypename