Docstrings

GPUArrays.@LocalMemory — Macro.

Creates a local static memory shared inside one block. Equivalent to __local of OpenCL or __shared__ (<variable>) of CUDA.

GPUArrays.@allowscalar

GPUArrays.@cartesianidx — Macro.

cartesianidx(A, statesym = :state)

Like @linearidx(A, statesym = :state), but returns an N-dimensional NTuple{ndim(A), Int} as index

GPUArrays.@disallowscalar

GPUArrays.@linearidx — Macro.

linearidx(A, statesym = :state)

Macro form of linear_index, which calls return when out of bounds. So it can be used like this:

```julia
function kernel(state, A)
    idx = @linear_index A state
    # from here on it's save to index into A with idx
    @inbounds begin
        A[idx] = ...
    end
end
```

GPUArrays.AbstractDeviceArray

GPUArrays.CommonReduceResult

GPUArrays.FFTKernel

GPUArrays.FFTPlan

GPUArrays.GPUArray

GPUArrays.GPUArrays

GPUArrays.GPUBackend

GPUArrays.GPUDestArray

GPUArrays.GPUMatrix

GPUArrays.GPUVecOrMat

GPUArrays.GPUVector

GPUArrays.JLArray

GPUArrays.JLBackend

GPUArrays.JLDevice

GPUArrays.JLState

GPUArrays.LCGStep

GPUArrays.LocalMem — Type.

Thread group local memory

GPUArrays.LocalMemory — Type.

Creates a block local array pointer with T being the element type and N the length. Both T and N need to be static! C is a counter for approriately get the correct Local mem id in CUDAnative. This is an internal method which needs to be overloaded by the GPU Array backends

GPUArrays.Sampler

GPUArrays.SmallSigned

GPUArrays.SmallUnsigned

GPUArrays.TausStep

GPUArrays.TestSuite

GPUArrays.WidenReduceResult

GPUArrays._allowscalar

GPUArrays._getindex

GPUArrays._gpu_call

GPUArrays._ind2sub

GPUArrays._setindex!

GPUArrays._sub2ind

GPUArrays.acc_mapreduce

GPUArrays.all_devices — Function.

Returns all devices from backends = active_backends(). Device can be filtered by passing filter_funcs, e.g. : is_gpu, is_cpu, (dev)-> has_atleast(dev, threads, 512)

GPUArrays.allequal

GPUArrays.allowscalar

GPUArrays.assertscalar

GPUArrays.available_devices — Function.

Returns all devices for the current backend. Device can be filtered by passing filter_funcs, e.g. : is_gpu, is_cpu, (dev)-> has_atleast(dev, threads, 512)

GPUArrays.backend

GPUArrays.bgetindex

GPUArrays.blas_module

GPUArrays.blasbuffer

GPUArrays.blockdim_x

GPUArrays.blockdim_y

GPUArrays.blockdim_z

GPUArrays.blockidx_x

GPUArrays.blockidx_y

GPUArrays.blockidx_z

GPUArrays.blocks — Function.

Blocks that group together hardware threads

GPUArrays.cached_state

GPUArrays.clear_cache

GPUArrays.collect_kernel

GPUArrays.convolution!

GPUArrays.convolution_fft!

GPUArrays.convolution_kernel

GPUArrays.copy_kernel!

GPUArrays.destroy! — Function.

Destroys context, freeing all it's resources.

GPUArrays.device — Function.

device(A::AbstractArray)

Gets the device associated to the Array A

GPUArrays.device_summary — Function.

Summarizes all features of a device and prints it to io

GPUArrays.eltype_or

GPUArrays.eval

GPUArrays.fast_isapprox — Function.

Same as Base.isapprox, but without keyword args and without nans

GPUArrays.fftkernel

GPUArrays.filterfuncs

GPUArrays.floattype

GPUArrays.forall_devices — Function.

Iterates through all available devices and calls f(context) after initializing the standard context for that device.

GPUArrays.free_global_memory — Function.

Free global memory. Isn't supported for AMD cards right now, in which case it returns NaN, so don't rely on the output of this function.

GPUArrays.genperm

GPUArrays.getidx_2d1d

GPUArrays.global_memory — Function.

Global memory, e.g. VRAM or RAM of device

GPUArrays.global_size — Function.

global_size(state)

Global size == blockdim * griddim == total number of kernel execution

GPUArrays.gpu_call — Function.

gpu_call(kernel::Function, A::GPUArray, args::Tuple, configuration = length(A))

Calls function kernel on the GPU. A must be an GPUArray and will help to dispatch to the correct GPU backend and supplies queues and contexts. Calls the kernel function with kernel(state, args...), where state is dependant on the backend and can be used for getting an index into A with linear_index(state). Optionally, a launch configuration can be supplied in the following way:

1) A single integer, indicating how many work items (total number of threads) you want to launch.
    in this case `linear_index(state)` will be a number in the range `1:configuration`
2) Pass a tuple of integer tuples to define blocks and threads per blocks!

GPUArrays.gpu_convert

GPUArrays.gpu_ind2sub

GPUArrays.gpu_promote_type

GPUArrays.gpu_rand

GPUArrays.gpu_sub2ind

GPUArrays.griddim_x

GPUArrays.griddim_y

GPUArrays.griddim_z

GPUArrays.grow_at

GPUArrays.grow_dimensions

GPUArrays.has_atleast — Function.

Checks a device for a certain attribute and returns true if it has at least value. Can be used with e.g. threads, blocks, global_memory, local_memory

GPUArrays.include

GPUArrays.index_kernel

GPUArrays.indexstyle

GPUArrays.is_cpu — Function.

Returns true if device is a cpu

GPUArrays.is_cudanative

GPUArrays.is_gpu — Function.

Returns true if device is a gpu

GPUArrays.is_julia

GPUArrays.is_opencl

GPUArrays.is_opengl

GPUArrays.linear_index — Function.

linear_index(state)

linear index corresponding to each kernel launch (in OpenCL equal to getglobalid).

GPUArrays.local_memory — Function.

Block local memory

GPUArrays.make_rand_num

GPUArrays.mapreduce_impl

GPUArrays.mapreducedim_kernel

GPUArrays.materialize

GPUArrays.name — Function.

Hardware name of a device

GPUArrays.new_context — Function.

Creates a new context from device without caching the resulting context.

GPUArrays.next_rand

GPUArrays.on_device — Function.

Creates a temporary context for device and executes f(context) while this context is active. Context gets destroyed afterwards. Note, that creating a temporary context is expensive.

GPUArrays.reduce_kernel

GPUArrays.reset! — Function.

Resets a context freeing all resources and creating a new context.

GPUArrays.reshape!

GPUArrays.setindex_kernel!

GPUArrays.shmem_counter

GPUArrays.simple_broadcast_index

GPUArrays.startvalue

GPUArrays.supports_double

GPUArrays.synchronize — Function.

synchronize(A::AbstractArray)

Blocks until all operations are finished on A

GPUArrays.synchronize_threads — Function.

 synchronize_threads(state)

in CUDA terms __synchronize in OpenCL terms: barrier(CLK_LOCAL_MEM_FENCE)

GPUArrays.test — Function.

Runs the entire GPUArrays test suite on array type AT

GPUArrays.thread_blocks_heuristic

GPUArrays.threadidx_x

GPUArrays.threadidx_y

GPUArrays.threadidx_z

GPUArrays.threads — Function.

Hardware threads of device

GPUArrays.to_blocks

GPUArrays.to_cartesian

GPUArrays.to_cpu

GPUArrays.to_device

GPUArrays.to_index

GPUArrays.to_number_range

GPUArrays.transpose_blocks!

GPUArrays.uniformscaling_kernel

GPUArrays.unsafe_reinterpret — Function.

Unsafe reinterpret for backends to overload. This makes it easier to do checks just on the high level.

GPUArrays.update! — Function.

This updates an array, even if dimensions and sizes don't match. Will resize accordingly!