Docstrings

TestImages.REPO_URL

TestImages.TestImages

TestImages.eval

TestImages.imagedir

TestImages.include

TestImages.remotefiles

TestImages.testimage — Function.

testimage(filename, [ops...])

load test image that partially matches filename, the first match is used if there're more than one. If ops is specified, it will be passed to load function. use TestImages.remotefiles to get a full list of available images.

Example

julia> using TestImages
julia> testimage("cameraman.tif")
julia> testimage("cameraman")
julia> testimage("c")