Docstrings

SplitApplyCombine.CombineDimsArray

SplitApplyCombine.GetFielder

SplitApplyCombine.Groups

SplitApplyCombine.Indexer

SplitApplyCombine.MappedArray

SplitApplyCombine.MappedIterator

SplitApplyCombine.ProductArray

SplitApplyCombine.SplitApplyCombine

SplitApplyCombine.SplitDimsArray

SplitApplyCombine._combinedims

SplitApplyCombine._eltypes

SplitApplyCombine._groupinds

SplitApplyCombine._inneraxes

SplitApplyCombine._innerjoin!

SplitApplyCombine._invert!

SplitApplyCombine._splitdims

SplitApplyCombine.check_dims

SplitApplyCombine.combinedims — Function.

combinedims(array_of_arrays)

Combine the dimensions of a nested array structure into a new, flat array.

This is the inverse operation of splitdims / splitdimsview.

See also combinedimsview, the lazy version of this function.

Example

julia> combinedims([[1, 2], [3, 4]])
2×2 Array{Int64,2}:
 1  3
 2  4

SplitApplyCombine.combinedimsview — Function.

combinedimsview(array_of_arrays)

Lazily create a flat array view of a nested array structure.

This is the inverse operation of splitdims / splitdimsview.

See also combinedims, the eager version of this function.

SplitApplyCombine.default_innerdims

SplitApplyCombine.eval

SplitApplyCombine.flatten — Function.

flatten(a)

Takes a collection of collections a and returns a collection containing all the elements of the subcollecitons of a. Equivalent to mapmany(idenity, a).

Example

julia> flatten([1:1, 1:2, 1:3])
6-element Array{Int64,1}:
 1
 1
 2
 1
 2
 3

SplitApplyCombine.group — Function.

group(by, iter)

Group the elements x of the iterable iter into groups labeled by by(x).

Example

julia> group(iseven, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10])
Dict{Bool,Array{Int64,1}} with 2 entries:
  false => [1, 3, 5, 7, 9]
  true  => [2, 4, 6, 8, 10]

group(by, f, iter...)

Sorts the elements x of the iterable iter into groups labeled by by(x), transforming each element to f(x). If multiple collections (of the same length) are provided, the transformations occur elementwise.

Example

julia> group(iseven, x -> x ÷ 2, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10])
Dict{Bool,Array{Int64,1}} with 2 entries:
  false => [0, 1, 2, 3, 4]
  true  => [1, 2, 3, 4, 5]

SplitApplyCombine.group2

SplitApplyCombine.groupcount — Function.

groupcount([by], iter)

Determine the number of elements of iter belonging to each group.

SplitApplyCombine.groupinds — Function.

groupinds(by, iter...)

Sorts the indices i of iter into groups labeled by by(iter[i]). If multiple collections (with matching indices) are provided, the groups are formed elementwise.

Example

julia> groupinds(iseven, [3,4,2,6,5,8])
Dict{Bool,Array{Int64,1}} with 2 entries:
  false => [1, 5]
  true  => [2, 3, 4, 6]

SplitApplyCombine.grouplength

SplitApplyCombine.groupprod — Function.

groupprod(by, [f], iter)

Multiply the elements of iter belonging to different groups, optionally mapping by f.

SplitApplyCombine.groupreduce — Function.

groupreduce(by, f, op, iter...; [init])

Applies a mapreduce-like operation on the groupings labeled by passing the elements of iter through by. Mostly equivalent to map(g -> reduce(op, g), group(by, f, iter)), but designed to be more efficient. If multiple collections (of the same length) are provided, the transformations by and f occur elementwise.

groupreduce(by, op, iter; [init])

Like groupreduce(by, identity, op, iter; init=init).

Example

julia> groupreduce(iseven, +, 1:10)
Dict{Bool,Int64} with 2 entries:
false => 25
true  => 30

SplitApplyCombine.groupsum — Function.

groupsum(by, [f], iter)

Sum the elements of iter belonging to different groups, optionally mapping by f.

SplitApplyCombine.groupview — Function.

groupview(by, iter)

Like group, but each grouping is a view of the input collection iter.

Example

julia> v = [3,4,2,6,5,8]
6-element Array{Int64,1}:
 3
 4
 2
 6
 5
 8

julia> groups = groupview(iseven, v)
SAC.Groups{Bool,Any,Array{Int64,1},Dict{Bool,Array{Int64,1}}} with 2 entries:
  false => [3, 5]
  true  => [4, 2, 6, 8]

julia> groups[false][:] = 99
99

julia> v
6-element Array{Int64,1}:
 99
  4
  2
  6
 99
  8

SplitApplyCombine.include

SplitApplyCombine.inner_eltype

SplitApplyCombine.innerjoin — Function.

innerjoin(lkey, rkey, f, comparison, left, right)

Performs a relational-style join operation between iterables left and right, returning a collection of elements f(l, r) for which comparison(lkey(l), rkey(r)) is true where l ∈ left, r ∈ right.

Example

julia> innerjoin(iseven, iseven, tuple, ==, [1,2,3,4], [0,1,2])
6-element Array{Tuple{Int64,Int64},1}:
 (1, 1)
 (2, 0)
 (2, 2)
 (3, 1)
 (4, 0)
 (4, 2)

SplitApplyCombine.innerjoin!

SplitApplyCombine.invert — Function.

invert(a)

Return a new nested container by reversing the order of the nested container a, for example turning a dictionary of arrays into an array of dictionaries, such that a[i][j] === invert(a)[j][i].

Note that in order for the keys of the inner and outer structure to be known, the input container a must not be empty.

Examples

julia> invert([[1,2,3],[4,5,6]])
3-element Array{Array{Int64,1},1}:
 [1, 4]
 [2, 5]
 [3, 6]

julia> invert((a = [1, 2, 3], b = [2.0, 4.0, 6.0]))
3-element Array{NamedTuple{(:a, :b),Tuple{Int64,Float64}},1}:
 (a = 1, b = 2.0)
 (a = 2, b = 4.0)
 (a = 3, b = 6.0)

SplitApplyCombine.invert!

SplitApplyCombine.leftgroupjoin — Function.

leftgroupjoin(lkey, rkey, f, comparison, left, right)

Creates a collection if groups labelled by lkey(l) where each group contains elements f(l, r) which satisfy comparison(lkey(l), rkey(r)). If there are no matches, the group is still created (with an empty collection).

This operation shares some similarities with an SQL left outer join.

Example

julia> leftgroupjoin(iseven, iseven, tuple, ==, [1,2,3,4], [0,1,2])
Dict{Bool,Array{Tuple{Int64,Int64},1}} with 2 entries:
  false => Tuple{Int64,Int64}[(1, 1), (3, 1)]
  true  => Tuple{Int64,Int64}[(2, 0), (2, 2), (4, 0), (4, 2)]

SplitApplyCombine.mapmany — Function.

mapmany(f, a...)

Like map, but f(x) for each x ∈ a may return an arbitrary number of values to insert into the output.

Example

julia> mapmany(x -> 1:x, [1,2,3])
6-element Array{Int64,1}:
 1
 1
 2
 1
 2
 3

SplitApplyCombine.mapview — Function.

mapview(f, a)

Return a view of a where each element is mapped through function f. The iteration and indexing properties of a are preserved. Similar to map(f, a), except evaluated lazily.

Example

julia> a = [1,2,3];

julia> b = mapview(-, a)
3-element MappedArray{Int64,1,typeof(-),Array{Int64,1}}:
 -1
 -2
 -3

julia> a[1] = 10;

julia> b
3-element MappedArray{Int64,1,typeof(-),Array{Int64,1}}:
 -10
  -2
  -3

SplitApplyCombine.new_eltype

SplitApplyCombine.product — Function.

product(f, a, b)

Takes the Cartesian outer product of two containers and evaluates f on all pairings of elements.

For example, if a and b are vectors, this returns a matrix out such that out[i,j] = f(a[i], b[j]) for i in keys(a) and j in keys(b). See also productview.

Example

julia> product(+, [1,2], [1,2,3])
2×3 Array{Int64,2}:
 2  3  4
 3  4  5

SplitApplyCombine.product!

SplitApplyCombine.productview — Function.

productview(f, a, b)

Return a view of a Cartesian product of a and b where the output elements are f evaluated those of a and b. See also product and ProductArray

Example

julia> productview(+, [1,2], [1,2,3])
2×3 ProductArray{Int64,2,typeof(+),Array{Int64,1},Array{Int64,1}}:
 2  3  4
 3  4  5

SplitApplyCombine.single

SplitApplyCombine.slice_inds

SplitApplyCombine.splat_inds

SplitApplyCombine.splitdims — Function.

splitdims(array, [dims])

Eagerly split the contents of array into a nested array of arrays. The outermost array contains the specified dimension(s) dims, which may be an integer, a tuple of integers, or defaults to the final array dimension. The nested arrays will contain all the remaining dimensions (in ascending order).

See also splitdimsview, which performs a similar operation lazily.

Examples:

julia> splitdims([1 2; 3 4])
2-element Array{Array{Int64,1},1}:
 [1, 3]
 [2, 4]

julia> splitdims([1 2; 3 4], 1)
2-element Array{Array{Int64,1},1}:
 [1, 2]
 [3, 4]

SplitApplyCombine.splitdimsview — Function.

splitdimsview(array, [dims])

Lazily split the contents of array into a nested array of arrays. The outermost array contains the specified dimension(s) dims, which may be an integer, a tuple of integers, or defaults to the final array dimension. The nested arrays will contain all the remaining dimensions (in ascending order).

See also splitdims, which performs a similar operation eagerly.

Examples:

julia> splitdims([1 2; 3 4])
2-element SplitDimsArray{SubArray{Int64,1,Array{Int64,2},Tuple{Base.Slice{Base.OneTo{Int64}},Int64},true},1,(2,),Array{Int64,2}}:
 [1, 3]
 [2, 4]

julia> splitdims([1 2; 3 4], 1)
2-element SplitDimsArray{SubArray{Int64,1,Array{Int64,2},Tuple{Int64,Base.Slice{Base.OneTo{Int64}}},true},1,(1,),Array{Int64,2}}:
 [1, 2]
 [3, 4]

SplitApplyCombine.⨝