Docstrings

WoodburyMatrices.SymWoodbury — Type.

Represents a matrix of the form A + BDBᵀ.

WoodburyMatrices.Woodbury — Type.

W = Woodbury(A, U, C, V) creates a matrix W identical to A + U*C*V whose inverse will be calculated using the Woodbury matrix identity.

WoodburyMatrices.WoodburyMatrices

WoodburyMatrices.calc_inv

WoodburyMatrices.conjm

WoodburyMatrices.eval

WoodburyMatrices.include

WoodburyMatrices.inv_invD_BtX

WoodburyMatrices.liftFactor — Function.

liftFactor(A)

More stable version of inv(A). Returns a function which computs the inverse on evaluation, i.e. liftFactor(A)(x) is the same as inv(A)*x.

WoodburyMatrices.liftFactorVars

WoodburyMatrices.partialInv — Function.

partialInv(O)

Get the factors (X,Z) in W + XZXᵀ where W + XZXᵀ = inv( A + BDBᵀ )

WoodburyMatrices.plusBDBtx!

WoodburyMatrices.sparse_factors — Function.

Given a type T, an integer n and m tuples of the form (i, j, v), build sparse matrices rows, vals, cols such that the product out = rows * vals * cols is equivalent to:

out = zeros(T, n, n)

for (i, j, v) in args
    out[i, j] = v
end

The first two components (i and j) of each tuple should be integers whereas the third component should be of type T

Example:

julia> r, v, c = WoodburyMatrices.sparse_factors(Float64, 3,
                                                 (1, 1, 2.0),
                                                 (2, 2, 3.0),
                                                 (3, 3, 4.0));

julia> r*c*v - Diagonal([2.0, 3.0, 4.0])
3x3 sparse matrix with 0 Float64 entries:

WoodburyMatrices.square