Docstrings

JuliaWebAPI.APIInvoker — Type.

APIInvoker holds the transport and format used for a remote api call.

JuliaWebAPI.APIResponder — Type.

APIResponder holds the transport and format used for data exchange and the endpoint specifications.

JuliaWebAPI.APISpec

JuliaWebAPI.AbstractAPIInvoker

JuliaWebAPI.AbstractAPIResponder

JuliaWebAPI.AbstractMsgFormat

JuliaWebAPI.AbstractTransport

JuliaWebAPI.CONTROL_CMDS

JuliaWebAPI.CRLFCRLF

JuliaWebAPI.DictMsgFormat — Type.

Intermediate format that is just a Dict. No serialization. A Dict with cmd (string), args (array), vargs (dict).

JuliaWebAPI.ERR_CODES

JuliaWebAPI.EndPts

JuliaWebAPI.HttpRpcServer

JuliaWebAPI.InProcChannels

JuliaWebAPI.InProcTransport — Type.

Transport layer over in-process Channels

JuliaWebAPI.JSONMsgFormat — Type.

Intermediate format based on JSON. A JSON object with cmd (string), args (array), vargs (dict).

JuliaWebAPI.JuliaWebAPI

JuliaWebAPI.LFLF — Constant.

Handles multipart form data. Transfers all content as String. Binary files can be uplaoded by encoding them with base64 first.

JuliaWebAPI.SerializedMsgFormat — Type.

Intermediate format based on Julia serialization. A dict with cmd (string), args (array), vargs (dict).

JuliaWebAPI.ZMQTransport — Type.

Transport layer over ZMQ sockets

JuliaWebAPI._add_spec

JuliaWebAPI._dict_dser

JuliaWebAPI._dict_fmt

JuliaWebAPI._dict_fnresponse

JuliaWebAPI._dict_httpresponse

JuliaWebAPI._dict_ser

JuliaWebAPI.apicall — Function.

Calls a remote api cmd with args... and data.... The response is formatted as specified by the formatter specified in conn.

JuliaWebAPI.args

JuliaWebAPI.call_api — Function.

call the actual API method, and send the return value back as response

JuliaWebAPI.close

JuliaWebAPI.cmd

JuliaWebAPI.collect_part_data

JuliaWebAPI.create_responder

JuliaWebAPI.data

JuliaWebAPI.data_dict

JuliaWebAPI.default_endpoint

JuliaWebAPI.default_preproc

JuliaWebAPI.dynamic_invoke

JuliaWebAPI.eval

JuliaWebAPI.fnresponse — Function.

extract and return the response data as a direct function call would have returned but throw error if the call was not successful.

JuliaWebAPI.get_hdrs

JuliaWebAPI.get_multipart_form_boundary

JuliaWebAPI.get_resp

JuliaWebAPI.header

JuliaWebAPI.http_handler

JuliaWebAPI.httpresponse — Function.

construct an HTTP Response object from the API response

JuliaWebAPI.include

JuliaWebAPI.isvalidcmd

JuliaWebAPI.juliaformat

JuliaWebAPI.logerr

JuliaWebAPI.make_vargs

JuliaWebAPI.narrow_args!

JuliaWebAPI.parse_part_headers

JuliaWebAPI.parsepostdata

JuliaWebAPI.process — Function.

start processing as a server

JuliaWebAPI.process_async

JuliaWebAPI.promote_arr

JuliaWebAPI.recvreq

JuliaWebAPI.register — Function.

Register a function as API call. TODO: validate method belongs to module?

JuliaWebAPI.respond — Function.

send a response over the transport in the specified format

JuliaWebAPI.run_http

JuliaWebAPI.run_rest

JuliaWebAPI.searcharr

JuliaWebAPI.sendrecv

JuliaWebAPI.sendresp

JuliaWebAPI.wireformat