Docstrings

Planets.Planets

Planets.accretion_rate — Function.

Compute the gas accretion rate onto a planet, up to Neptune size, embedded in a protoplanetary disk. This function implements Equation (B36) derived in Carrera et al. (2018) which itself is adapted from Ginzburg et al. (2016).

NOTE: Equation (B36) of Carrera et al. is in units of Mearth/Myr but this function returns values in untis of Mearth/year.

NOTE: The accretion rate for a planet with zero atmosphere diverges. You need to initialize the planet's atmosphere to a non-zero value. A reasonable choice is the gas mass contained inside the planet's Bondi radius. Furthermore, the accretion rate is initially very high and small timesteps are required to resolve the accretion correctly.

Example:

#
# Gas accretion rate (units: M_earth / year) for a small planet embedded
# in a protoplanetary disk.
#
# Inputs:
#
M_core = 8.0  # Planet's core mass in Earth masses.
M_atm = 0.02  # Planet's current H2 mass in Earth masses.
T_disk = 500  # Local disk temperature in Kelvin.

M_atm_dot = accretion_rate(M_core, M_atm, T_disk) # M_earth / year.

Citation:

Carrera et al. (2018) and Ginzburg et al. (2016)

http://adsabs.harvard.edu/cgi-bin/bib_query?arXiv:1804.05069
http://adsabs.harvard.edu/abs/2016ApJ...825...29G

Initializing the planet's atmosphere:

ρ = ... # Local disk gas density in g/cm^3
T = ... # Local disk gas temperature in Kelvin

M_core = ... # Core mass in M_earth

Rb = 2.861e7 * M_core / T # Bondi radius in cm

M_atm = ρ * Rb^3 * 1.6744e-28 # Initial atmosphere mass in M_earth

Planets.core_radius — Function.

Compute the core radius for either a rock-iron or water-rock core.

Examples:

#
# Radius of a 3.0 M_earth core with 10% water + 90% rock:
#
radius = core_radius(3.0, h2o=0.1) # In Earth radii.

#
# Radius of a 3.0 M_earth core with 10% iron + 90% rock:
#
radius = core_radius(3.0, fe=0.1) # In Earth radii.

Citation:

Zeng et al (2016)
http://adsabs.harvard.edu/abs/2016ApJ...819..127Z

Planets.eval

Planets.habitable_zone — Function.

Computes all the limits for the habitable zone from Kopparapu et al. (2013). This function uses the updated coefficients from their website.

Input:

Teff       Star's effective temperature (K)
L_star     Star's luminosity (L_sun)

Example:

Teff   = 3700  # Temperature of a 0.5 M_sun star.
L_star = 0.04  # Luminosity of a 0.5 M_sun star.

limits = habitable_zone(Teff, L_star)

@info("Recent Venus                  = $(limits[1]) AU")
@info("Runaway Greenhouse            = $(limits[2]) AU")
@info("Maximum Greenhouse            = $(limits[3]) AU")
@info("Early Mars                    = $(limits[4]) AU")
@info("Runaway Greenhouse for 5.0 ME = $(limits[5]) AU")
@info("Runaway Greenhouse for 0.1 ME = $(limits[6]) AU")

Citations:

Kopparapu et al. (2013)
http://adsabs.harvard.edu/abs/2013ApJ...765..131K

Erratum
http://adsabs.harvard.edu/abs/2013ApJ...770...82K

Updated values
https://depts.washington.edu/naivpl/sites/default/files/hz.shtml

Planets.include

Planets.interpolate_mass

Planets.interpolate_time_and_mass

Planets.load_parsec — Function.

This is an internal function (not exported). It is used to load one of the stellar evolution tables from the data/ directory. The caller is responsible for supplying a valid table name:

"Z05"	,	"Z10"	,	"Z15"	,	"Z20"

The data is saved as a data frame inside the global variable parsec[...]. If the data was previously loaded, it won't be loaded again, so it's safe to call this function many times.

Example:

load_parsec("Z05")

... now parsec["Z05"] contains data.

Planets.mr_names

Planets.mr_table

Planets.parsec

Planets.stellar_evolution — Function.

Compute the stellar evolution tracks (luminosity, temperature, logg) for AFGKM with a metallicity of Z = 1.0%, up to an age of 0.89 Gyr, using the stellar models of Marigo et al. (2017). This function the following values:

Teff     Star's effective temperature (K)
L_star   Log base 10 of the stellar luminosity (L_sun)
logg     Log gravity.

Example:

M_star = 0.5 # Stellar mass (M_sun)
age = 1e9    # Stellar age (years)
Z = 0.01     # Metallicity

Teff, L_star, logg = stellar_evolution(M_star,t=age,Z=Z)

Citation:

Marigo et al. (2017)
http://stev.oapd.inaf.it/cgi-bin/cmd
http://adsabs.harvard.edu/abs/2017ApJ...835...77M