Docstrings

DataFramesMeta.@based_on — Macro.

@based_on(g, i...)

Summarize a grouping operation

Arguments

g : a GroupedDataFrame

i... : keyword arguments defining new columns

Examples

julia> using DataFramesMeta, DataFrames

julia> d = DataFrame(
            n = 1:20,
            x = [3, 3, 3, 3, 1, 1, 1, 2, 1, 1, 2, 1, 1, 2, 2, 2, 3, 1, 1, 2]);

julia> g = groupby(d, :x);

julia> @based_on(g, nsum = sum(:n))
3×2 DataFrames.DataFrame
│ Row │ x │ nsum │
├─────┼───┼──────┤
│ 1   │ 1 │ 99   │
│ 2   │ 2 │ 84   │
│ 3   │ 3 │ 27   │

julia> @based_on(g, x2 = 2 * :x, nsum = sum(:n))
20×3 DataFrames.DataFrame
│ Row │ x │ x2 │ nsum │
├─────┼───┼────┼──────┤
│ 1   │ 1 │ 2  │ 99   │
│ 2   │ 1 │ 2  │ 99   │
│ 3   │ 1 │ 2  │ 99   │
│ 4   │ 1 │ 2  │ 99   │
│ 5   │ 1 │ 2  │ 99   │
│ 6   │ 1 │ 2  │ 99   │
│ 7   │ 1 │ 2  │ 99   │
│ 8   │ 1 │ 2  │ 99   │
⋮
│ 12  │ 2 │ 4  │ 84   │
│ 13  │ 2 │ 4  │ 84   │
│ 14  │ 2 │ 4  │ 84   │
│ 15  │ 2 │ 4  │ 84   │
│ 16  │ 3 │ 6  │ 27   │
│ 17  │ 3 │ 6  │ 27   │
│ 18  │ 3 │ 6  │ 27   │
│ 19  │ 3 │ 6  │ 27   │
│ 20  │ 3 │ 6  │ 27   │

DataFramesMeta.@by — Macro.

@by(d::AbstractDataFrame, cols, e...)

Split-apply-combine in one step.

Arguments

d : an AbstractDataFrame

cols : a column indicator (Symbol, Int, Vector{Symbol}, etc.)

e : keyword arguments specifying new columns in terms of column groupings

Returns

::DataFrame

Examples

julia> using DataFrames, DataFramesMeta, Statistics

julia> df = DataFrame(
            a = repeat(1:4, outer = 2),
            b = repeat(2:-1:1, outer = 4),
            c = randn(8));

julia> @by(df, :a, d = sum(:c))
4×2 DataFrames.DataFrame
│ Row │ a │ d        │
├─────┼───┼──────────┤
│ 1   │ 1 │ 1.27638  │
│ 2   │ 2 │ 1.00951  │
│ 3   │ 3 │ 1.48328  │
│ 4   │ 4 │ -2.42621 │

julia> @by(df, :a, d = 2 * :c)
8×2 DataFrames.DataFrame
│ Row │ a │ d         │
├─────┼───┼───────────┤
│ 1   │ 1 │ 1.22982   │
│ 2   │ 1 │ 1.32294   │
│ 3   │ 2 │ 1.93664   │
│ 4   │ 2 │ 0.0823819 │
│ 5   │ 3 │ -0.670512 │
│ 6   │ 3 │ 3.63708   │
│ 7   │ 4 │ -3.06436  │
│ 8   │ 4 │ -1.78806  │

julia> @by(df, :a, c_sum = sum(:c), c_mean = mean(:c))
4×3 DataFrames.DataFrame
│ Row │ a │ c_sum    │ c_mean   │
├─────┼───┼──────────┼──────────┤
│ 1   │ 1 │ 1.27638  │ 0.63819  │
│ 2   │ 2 │ 1.00951  │ 0.504755 │
│ 3   │ 3 │ 1.48328  │ 0.741642 │
│ 4   │ 4 │ -2.42621 │ -1.2131  │

julia> @by(df, :a, c = :c, c_mean = mean(:c))
8×3 DataFrames.DataFrame
│ Row │ a │ c         │ c_mean   │
├─────┼───┼───────────┼──────────┤
│ 1   │ 1 │ 0.61491   │ 0.63819  │
│ 2   │ 1 │ 0.66147   │ 0.63819  │
│ 3   │ 2 │ 0.968319  │ 0.504755 │
│ 4   │ 2 │ 0.041191  │ 0.504755 │
│ 5   │ 3 │ -0.335256 │ 0.741642 │
│ 6   │ 3 │ 1.81854   │ 0.741642 │
│ 7   │ 4 │ -1.53218  │ -1.2131  │
│ 8   │ 4 │ -0.894029 │ -1.2131  │

DataFramesMeta.@byrow! — Macro.

@byrow!(d, expr)

Act on a DataFrame row-by-row.

Includes support for control flow and begin end blocks. Since the "environment" induced by @byrow! df is implicitly a single row of df, use regular operators and comparisons instead of their elementwise counterparts as in @with. Note that the scope within @byrow! is a hard scope.

byrow! also supports special syntax for allocating new columns. The syntax @newcol x::Array{Int} allocates a new column :x with an Array container with eltype Int. Note that the returned AbstractDataFrame includes these new columns, but the original d is not affected. This feature makes it easier to use byrow! for data transformations. _N is introduced to represent the length of the dataframe, _D represents the dataframe including added columns, and row represents the index of the current row.

Arguments

d : an AbstractDataFrame

expr : expression operated on row by row

Returns

The modified AbstractDataFrame.

Examples

julia> using DataFrames, DataFramesMeta

julia> df = DataFrame(A = 1:3, B = [2, 1, 2]);

julia> let x = 0
            @byrow! df begin
                if :A + :B == 3
                    x += 1
                end
            end  #  This doesn't work without the let
            x
        end
2

julia> @byrow! df begin
            if :A > :B
                :A = 0
            end
        end
3×2 DataFrames.DataFrame
│ Row │ A │ B │
├─────┼───┼───┤
│ 1   │ 1 │ 2 │
│ 2   │ 0 │ 1 │
│ 3   │ 0 │ 2 │

julia> df2 = @byrow! df begin
           @newcol colX::Array{Float64}
           :colX = :B == 2 ? pi * :A : :B
       end
3×3 DataFrames.DataFrame
│ Row │ A │ B │ colX    │
├─────┼───┼───┼─────────┤
│ 1   │ 1 │ 2 │ 3.14159 │
│ 2   │ 0 │ 1 │ 1.0     │
│ 3   │ 0 │ 2 │ 0.0     │

DataFramesMeta.@linq — Macro.

@linq df ...

General macro that creates a mini DSL for chaining and macro calls.

Details

The following embedded function calls are equivalent to their macro version:

with

where

select

transform

by

groupby

orderby

based_on

Examples

julia> using DataFrames, DataFramesMeta, Statistics

julia> n = 100;

julia> df = DataFrame(a = rand(1:3, n),
                      b = ["a","b","c","d"][rand(1:4, n)],
                      x = rand(n));

julia> x1 = @linq transform(where(df, :a .> 2, :b .!= "c"), y = 10 * :x);

julia> x1 = @linq by(x1, :b, meanX = mean(:x), meanY = mean(:y));

julia> @linq select(orderby(x1, :b, -:meanX), var = :b, :meanX, :meanY)
3×3 DataFrames.DataFrame
│ Row │ var │ meanX    │ meanY   │
├─────┼─────┼──────────┼─────────┤
│ 1   │ "a" │ 0.665682 │ 6.65682 │
│ 2   │ "b" │ 0.617848 │ 6.17848 │
│ 3   │ "d" │ 0.568289 │ 5.68289 │

julia> @linq df |>
           transform(y = 10 * :x) |>
           where(:a .> 2) |>
           by(:b, meanX = mean(:x), meanY = mean(:y)) |>
           orderby(:meanX) |>
           select(:meanX, :meanY, var = :b)
4×3 DataFrames.DataFrame
│ Row │ meanX    │ meanY   │ var │
├─────┼──────────┼─────────┼─────┤
│ 1   │ 0.353205 │ 3.53205 │ "a" │
│ 2   │ 0.419833 │ 4.19833 │ "d" │
│ 3   │ 0.452061 │ 4.52061 │ "c" │
│ 4   │ 0.519316 │ 5.19316 │ "b" │

DataFramesMeta.@orderby — Macro.

@orderby(d, i...)

Sort by criteria. Normally used to sort groups in GroupedDataFrames.

Arguments

d : an AbstractDataFrame or GroupedDataFrame

i... : expression for sorting

Examples

julia> using DataFrames, DataFramesMeta, Statistics

julia> d = DataFrame(n = 1:20, x = [3, 3, 3, 3, 1, 1, 1, 2, 1, 1,
                                    2, 1, 1, 2, 2, 2, 3, 1, 1, 2]);

julia> g = groupby(d, :x);

julia> @orderby(g, mean(:n))
DataFrames.GroupedDataFrame  3 groups with keys: Symbol[:x]
First Group:
5×2 DataFrames.SubDataFrame{Array{Int64,1}}
│ Row │ n  │ x │
├─────┼────┼───┤
│ 1   │ 1  │ 3 │
│ 2   │ 2  │ 3 │
│ 3   │ 3  │ 3 │
│ 4   │ 4  │ 3 │
│ 5   │ 17 │ 3 │
⋮
Last Group:
6×2 DataFrames.SubDataFrame{Array{Int64,1}}
│ Row │ n  │ x │
├─────┼────┼───┤
│ 1   │ 8  │ 2 │
│ 2   │ 11 │ 2 │
│ 3   │ 14 │ 2 │
│ 4   │ 15 │ 2 │
│ 5   │ 16 │ 2 │
│ 6   │ 20 │ 2 │

DataFramesMeta.@select — Macro.

@select(d, e...)

Select and transform columns.

Arguments

d : an AbstractDataFrame or AbstractDict

e : keyword arguments specifying new columns in terms of existing columns or symbols to specify existing columns

Returns

::AbstractDataFrame or ::AbstractDict

Examples

julia> using DataFrames, DataFramesMeta

julia> d = Dict(:s => 3, :y => 44, :d => 5);

julia> @select(d, x = :y + :d, :s)
Dict{Symbol,Int64} with 2 entries:
  :s => 3
  :x => 49

julia> df = DataFrame(a = repeat(1:4, outer = 2), b = repeat(2:-1:1, outer = 4), c = randn(8))
8×3 DataFrames.DataFrame
│ Row │ a │ b │ c         │
├─────┼───┼───┼───────────┤
│ 1   │ 1 │ 2 │ -0.354685 │
│ 2   │ 2 │ 1 │ 0.287631  │
│ 3   │ 3 │ 2 │ -0.918007 │
│ 4   │ 4 │ 1 │ -0.352519 │
│ 5   │ 1 │ 2 │ 0.743501  │
│ 6   │ 2 │ 1 │ -1.27415  │
│ 7   │ 3 │ 2 │ 0.258456  │
│ 8   │ 4 │ 1 │ -0.460486 │

julia> @select(df, :c, :a)
8×2 DataFrames.DataFrame
│ Row │ c         │ a │
├─────┼───────────┼───┤
│ 1   │ -0.354685 │ 1 │
│ 2   │ 0.287631  │ 2 │
│ 3   │ -0.918007 │ 3 │
│ 4   │ -0.352519 │ 4 │
│ 5   │ 0.743501  │ 1 │
│ 6   │ -1.27415  │ 2 │
│ 7   │ 0.258456  │ 3 │
│ 8   │ -0.460486 │ 4 │

julia> @select(df, :c, x = :b + :c)
8×2 DataFrames.DataFrame
│ Row │ c         │ x         │
├─────┼───────────┼───────────┤
│ 1   │ -0.354685 │ 1.64531   │
│ 2   │ 0.287631  │ 1.28763   │
│ 3   │ -0.918007 │ 1.08199   │
│ 4   │ -0.352519 │ 0.647481  │
│ 5   │ 0.743501  │ 2.7435    │
│ 6   │ -1.27415  │ -0.274145 │
│ 7   │ 0.258456  │ 2.25846   │
│ 8   │ -0.460486 │ 0.539514  │

DataFramesMeta.@transform — Macro.

@transform(d, i...)

Add additional columns or keys based on keyword arguments.

Arguments

d : an AbstractDict type, AbstractDataFrame, or GroupedDataFrame

i... : keyword arguments defining new columns or keys

For AbstractDict types, @transform only works with keys that are symbols.

Returns

::AbstractDataFrame, ::AbstractDict, or ::GroupedDataFrame

Examples

julia> using DataFramesMeta, DataFrames

julia> d = Dict(:s => 3, :y => 44, :d => 5);

julia> @transform(d, x = :y + :d)
Dict{Symbol,Int64} with 4 entries:
  :d => 5
  :s => 3
  :y => 44
  :x => 49

julia> df = DataFrame(A = 1:3, B = [2, 1, 2]);

julia> @transform(df, a = 2 * :A, x = :A .+ :B)
3×4 DataFrames.DataFrame
│ Row │ A │ B │ a │ x │
├─────┼───┼───┼───┼───┤
│ 1   │ 1 │ 2 │ 2 │ 3 │
│ 2   │ 2 │ 1 │ 4 │ 3 │
│ 3   │ 3 │ 2 │ 6 │ 5 │

DataFramesMeta.@where — Macro.

@where(d, i...)

Select row subsets in AbstractDataFrames or groups in GroupedDataFrames.

Arguments

d : an AbstractDataFrame or GroupedDataFrame

i... : expression for selecting rows

Multiple i expressions are "and-ed" together.

Examples

julia> using DataFramesMeta, DataFrames

julia> df = DataFrame(x = 1:3, y = [2, 1, 2]);

julia> x = [2, 1, 0];

julia> @where(df, :x .> 1)
2×2 DataFrames.DataFrame
│ Row │ x │ y │
├─────┼───┼───┤
│ 1   │ 2 │ 1 │
│ 2   │ 3 │ 2 │

julia> @where(df, :x .> x)
2×2 DataFrames.DataFrame
│ Row │ x │ y │
├─────┼───┼───┤
│ 1   │ 2 │ 1 │
│ 2   │ 3 │ 2 │

julia> @where(df, :x .> x, :y .== 3)
0×2 DataFrames.DataFrame

julia> d = DataFrame(n = 1:20, x = [3, 3, 3, 3, 1, 1, 1, 2, 1, 1,
                                    2, 1, 1, 2, 2, 2, 3, 1, 1, 2]);

julia> g = groupby(d, :x);

julia> @where(d, :x .== 3)
5×2 DataFrames.DataFrame
│ Row │ n  │ x │
├─────┼────┼───┤
│ 1   │ 1  │ 3 │
│ 2   │ 2  │ 3 │
│ 3   │ 3  │ 3 │
│ 4   │ 4  │ 3 │
│ 5   │ 17 │ 3 │

julia> @where(g, length(:x) > 5)   # pick out some groups
DataFrames.GroupedDataFrame  2 groups with keys: Symbol[:x]
First Group:
9×2 DataFrames.SubDataFrame{Array{Int64,1}}
│ Row │ n  │ x │
├─────┼────┼───┤
│ 1   │ 5  │ 1 │
│ 2   │ 6  │ 1 │
│ 3   │ 7  │ 1 │
│ 4   │ 9  │ 1 │
│ 5   │ 10 │ 1 │
│ 6   │ 12 │ 1 │
│ 7   │ 13 │ 1 │
│ 8   │ 18 │ 1 │
│ 9   │ 19 │ 1 │
⋮
Last Group:
6×2 DataFrames.SubDataFrame{Array{Int64,1}}
│ Row │ n  │ x │
├─────┼────┼───┤
│ 1   │ 8  │ 2 │
│ 2   │ 11 │ 2 │
│ 3   │ 14 │ 2 │
│ 4   │ 15 │ 2 │
│ 5   │ 16 │ 2 │
│ 6   │ 20 │ 2 │

DataFramesMeta.@with — Macro.

@with(d, expr)

@with allows DataFrame columns or AbstractDict keys to be referenced as symbols.

Arguments

d : an AbstractDataFrame or AbstractDict type

expr : the expression to evaluate in d

Details

@with works by parsing the expression body for all columns indicated by symbols (e.g. :colA). Then, a function is created that wraps the body and passes the columns as function arguments. This function is then called. Operations are efficient because:

A pseudo-anonymous function is defined, so types are stable.

Columns are passed as references, eliminating DataFrame indexing.

The following

@with(d, :a .+ :b .+ 1)

becomes

tempfun(a, b) = a .+ b .+ 1
tempfun(d[:a], d[:b])

All of the other DataFramesMeta macros are based on @with.

If an expression is wrapped in ^(expr), expr gets passed through untouched. If an expression is wrapped in cols(expr), the column is referenced by the variable expr rather than a symbol.

Examples

julia> using DataFramesMeta

julia> y = 3;

julia> d = Dict(:s => 3, :y => 44, :d => 5);

julia> @with(d, :s + :y + y)
50

julia> df = DataFrame(x = 1:3, y = [2, 1, 2]);

julia> x = [2, 1, 0];

julia> @with(df, :y .+ 1)
3-element Array{Int64,1}:
 3
 2
 3

julia> @with(df, :x + x)
3-element Array{Int64,1}:
 3
 3
 3

julia> @with df begin
            res = 0.0
            for i in 1:length(:x)
                res += :x[i] * :y[i]
            end
            res
        end
10.0

julia> @with(df, df[:x .> 1, ^(:y)]) # The ^ means leave the :y alone
2-element Array{Int64,1}:
 1
 2

julia> colref = :x;

julia> @with(df, :y + cols(colref)) # Equivalent to df[:y] + df[colref]
 3
 3
 5

@with creates a function, so scope within @with is a local scope. Variables in the parent can be read. Writing to variables in the parent scope differs depending on the type of scope of the parent. If the parent scope is a global scope, then a variable cannot be assigned without using the global keyword. If the parent scope is a local scope (inside a function or let block for example), the global keyword is not needed to assign to that parent scope.

DataFramesMeta.DataFramesMeta

DataFramesMeta.SymbolParameter

DataFramesMeta._transform!

DataFramesMeta.addkey!

DataFramesMeta.and

DataFramesMeta.based_on_helper

DataFramesMeta.by_helper

DataFramesMeta.byrow_find_newcols

DataFramesMeta.byrow_helper

DataFramesMeta.byrow_replace

DataFramesMeta.eval

DataFramesMeta.expandargs

DataFramesMeta.include

DataFramesMeta.linq

DataFramesMeta.mapexpr

DataFramesMeta.onearg

DataFramesMeta.orderby

DataFramesMeta.orderby_helper

DataFramesMeta.orderbyconstructor

DataFramesMeta.protect_replace_syms!

DataFramesMeta.replace_dotted!

DataFramesMeta.replace_equals_with_kw

DataFramesMeta.replace_syms!

DataFramesMeta.replacechains

DataFramesMeta.replacefuns

DataFramesMeta.select

DataFramesMeta.select_helper

DataFramesMeta.transform

DataFramesMeta.transform_helper

DataFramesMeta.where

DataFramesMeta.where_helper

DataFramesMeta.with_anonymous

DataFramesMeta.with_helper