Docstrings

SingularIntegralEquations.AbstractNonlocalCalculusOperator

SingularIntegralEquations.Cauchy

SingularIntegralEquations.CauchyWeight

SingularIntegralEquations.ComplexPlane

SingularIntegralEquations.ConcreteConvolution

SingularIntegralEquations.ConcreteHilbert

SingularIntegralEquations.ConcreteNonlocalLaplacian

SingularIntegralEquations.ConcretePseudoHilbert

SingularIntegralEquations.ConcreteSingularIntegral

SingularIntegralEquations.Convolution

SingularIntegralEquations.DefaultHilbert

SingularIntegralEquations.DefaultPseudoHilbert

SingularIntegralEquations.DefaultSingularIntegral

SingularIntegralEquations.Directed — Type.

Directed represents a number that is a limit from either left (s=true) or right (s=false) For functions with branch cuts, it is assumed that the value is on the branch cut, Therefore not requiring tolerances. This will naturally give the analytic continuation.

SingularIntegralEquations.FundamentalSolutions

SingularIntegralEquations.GreensFun

SingularIntegralEquations.Hilbert

SingularIntegralEquations.HilbertWrapper

SingularIntegralEquations.HornerFunctional

SingularIntegralEquations.HypergeometricFunctions — Module.

This module calculates (generalized) hypergeometric functions:

mFn(a;b;z) = Σ_{k=0}^∞ (a_1,k) ⋯ (a_m,k) / (b_1,k) ⋯ (b_n,k) zᵏ/k!

SingularIntegralEquations.JacobiQ

SingularIntegralEquations.JacobiQWeight — Type.

JacobiQWeight weights a basis on ℂ\𝕀 weighted by (z+1)^α*(z-1)^β. Note the inconsistency of the parameters with JacobiQ. When the domain is [a,b] the weight is inferred by mapping to [-1,1]

SingularIntegralEquations.LegendreQ

SingularIntegralEquations.LogKernelAsymptotics

SingularIntegralEquations.LowRankIntegralOperator

SingularIntegralEquations.NonlocalLaplacian

SingularIntegralEquations.OffHilbert

SingularIntegralEquations.OffSingularIntegral

SingularIntegralEquations.PseudoHilbert

SingularIntegralEquations.PseudoHilbertWrapper

SingularIntegralEquations.SingularIntegral

SingularIntegralEquations.SingularIntegralEquations

SingularIntegralEquations.SingularIntegralWrapper

SingularIntegralEquations.Stieltjes

SingularIntegralEquations.WeightedJacobiQ

SingularIntegralEquations.HypergeometricFunctions._₂F₁ — Function.

Compute the Gauss hypergeometric function ₂F₁(a,b;c;z).

SingularIntegralEquations.HypergeometricFunctions._₃F₂ — Function.

Compute the generalized hypergeometric function ₃F₂(a₁,a₂,a₃;b₁,b₂;z) with the Maclaurin series.

SingularIntegralEquations.cauchy

SingularIntegralEquations.cauchycircleS

SingularIntegralEquations.cauchyintegral

SingularIntegralEquations.cauchymoment

SingularIntegralEquations.cauchyweight

SingularIntegralEquations.centroid

SingularIntegralEquations.convolutionProductFun

SingularIntegralEquations.csingularintegral

SingularIntegralEquations.default_OffHilbert

SingularIntegralEquations.default_OffSingularIntegral

SingularIntegralEquations.diam

SingularIntegralEquations.diam2

SingularIntegralEquations.disjoint_cauchy

SingularIntegralEquations.dist

SingularIntegralEquations.dist2

SingularIntegralEquations.divkhornersum

SingularIntegralEquations.drummond

SingularIntegralEquations.eval

SingularIntegralEquations.exterior_cauchy

SingularIntegralEquations.extrema2

SingularIntegralEquations.forwardsubstitution

SingularIntegralEquations.forwardsubstitution!

SingularIntegralEquations.fourier_lambda

SingularIntegralEquations.fourier_lambda_large

SingularIntegralEquations.fourier_lambda_small

SingularIntegralEquations.gradient

SingularIntegralEquations.greensfun_checkdomains

SingularIntegralEquations.greensfun_checkgreensfun

SingularIntegralEquations.haswrapperstructure

SingularIntegralEquations.hilbert

SingularIntegralEquations.hilbertforward

SingularIntegralEquations.hilbertmoment

SingularIntegralEquations.hornersum

SingularIntegralEquations.hornervector

SingularIntegralEquations.include

SingularIntegralEquations.integratejin

SingularIntegralEquations.interior_cauchy

SingularIntegralEquations.iprocess!

SingularIntegralEquations.iskewtransform!

SingularIntegralEquations.istieltjes

SingularIntegralEquations.jacobiQweight

SingularIntegralEquations.joukowskyinverse

SingularIntegralEquations.joukowskyinverseabs

SingularIntegralEquations.joukowskyinversereal

SingularIntegralEquations.kernels

SingularIntegralEquations.lhelm_riemann

SingularIntegralEquations.lhelmfs

SingularIntegralEquations.linesumstieltjesintegral

SingularIntegralEquations.logabs

SingularIntegralEquations.logabsjacobimoment

SingularIntegralEquations.logabslegendremoment

SingularIntegralEquations.logjacobimoment

SingularIntegralEquations.logkernel

SingularIntegralEquations.lommelS2

SingularIntegralEquations.HypergeometricFunctions.mFn — Function.

Compute the generalized hypergeometric function mFn(a;b;z) with the Maclaurin series.

SingularIntegralEquations.maxspace_rule

SingularIntegralEquations.multif

SingularIntegralEquations.normalderivative

SingularIntegralEquations.normalization

SingularIntegralEquations.orientation

SingularIntegralEquations.orientationsign

SingularIntegralEquations.process!

SingularIntegralEquations.pseudocauchy

SingularIntegralEquations.pseudohilbert

SingularIntegralEquations.pseudostieltjes

SingularIntegralEquations.realdivkhornersum

SingularIntegralEquations.realintegratejin

SingularIntegralEquations.singularintegral

SingularIntegralEquations.skewProductFun

SingularIntegralEquations.skewpoints

SingularIntegralEquations.skewtransform!

SingularIntegralEquations.slices

SingularIntegralEquations.sqrtabs

SingularIntegralEquations.sqrtx2

SingularIntegralEquations.sqrtx2abs

SingularIntegralEquations.sqrtx2real

SingularIntegralEquations.stieltjes

SingularIntegralEquations.stieltjesbackward

SingularIntegralEquations.stieltjesbackward!

SingularIntegralEquations.stieltjesforward

SingularIntegralEquations.stieltjesforward!

SingularIntegralEquations.stieltjesintegral

SingularIntegralEquations.stieltjesintervalrecurrence

SingularIntegralEquations.stieltjesjacobimoment

SingularIntegralEquations.stieltjeslegendremoment

SingularIntegralEquations.stieltjesmoment

SingularIntegralEquations.stieltjesmoment!

SingularIntegralEquations.superscripts

SingularIntegralEquations.testsieeval

SingularIntegralEquations.testsieoperators

SingularIntegralEquations.testsies

SingularIntegralEquations.transformtimes

SingularIntegralEquations.undirected

SingularIntegralEquations.union_rule

SingularIntegralEquations.x̄sqrtx2real

SingularIntegralEquations.⁺

SingularIntegralEquations.⁻

SingularIntegralEquations.ℂ

SingularIntegralEquations.∇n