Docstrings

LabelledArrays.@LArray — Macro.

@LArray Eltype Size Names
@LArray Values Names

Creates an LArray with names determined from the Names vector and values determined from the Values array. Otherwise, and eltype and size are used to make an LArray with undefined values.

For example:

a = @LArray Float64 (2,2) (:a,:b,:c,:d)
b = @LArray [1,2,3] (:a,:b,:c)

LabelledArrays.@LVector — Macro.

@LVector Type Names

Creates an LArray of dimension 1 with eltype and undefined values. Length is via the number of names given.

For example:

b = @LVector [1,2,3] (:a,:b,:c)

LabelledArrays.@SLArray — Macro.

@SLArray Size Names
@SLArray Eltype Size Names

Creates an anonymous function that builds a labelled static vector with eltype ElType with names determined from the Names with size Size. If no eltype is given, then the eltype is determined from the arguments in the constructor.

For example:

ABC = @SLArray (2,2) (:a,:b,:c,:d)
x = ABC(1.0,2.5,3.0,5.0)
x.a == 1.0
x.b == 2.5
x.c == x[3]
x.d == x[2,2]

LabelledArrays.@SLVector — Macro.

@SLVector Names
@SLVector Eltype Names

Creates an anonymous function that builds a labelled static vector with eltype ElementType with names determined from the Names. If no eltype is given, then the eltype is determined from the values in the constructor.

For example:

ABC = @SLVector (:a,:b,:c)
x = ABC(1.0,2.5,3.0)
x.a == 1.0
x.b == 2.5
x.c == x[3]

LabelledArrays.DisplayCell

LabelledArrays.LAStyle

LabelledArrays.LArray

LabelledArrays.LVector

LabelledArrays.LabelledArrays

LabelledArrays.SLArray

LabelledArrays.SLVector

LabelledArrays.eval

LabelledArrays.include

LabelledArrays.symnames