Docstrings

ShiftedArrays.CircShiftedArray — Type.

CircShiftedArray(parent::AbstractArray, shifts)

Custom AbstractArray object to store an AbstractArray parent circularly shifted by shifts steps (where shifts is a Tuple with one shift value per dimension of parent). Use copy to collect the values of a CircShiftedArray into a normal Array.

Examples

julia> v = [1, 3, 5, 4];

julia> s = CircShiftedArray(v, (1,))
4-element ShiftedArrays.CircShiftedArray{Int64,1,Array{Int64,1}}:
 4
 1
 3
 5

julia> copy(s)
4-element Array{Int64,1}:
 4
 1
 3
 5

ShiftedArrays.CircShiftedVector — Type.

CircShiftedVector{T, S<:AbstractArray}

Shorthand for CircShiftedArray{T, 1, S}.

ShiftedArrays.ShiftedArray — Type.

ShiftedArray(parent::AbstractArray, shifts, default)

Custom AbstractArray object to store an AbstractArray parent shifted by shifts steps (where shifts is a Tuple with one shift value per dimension of parent). For s::ShiftedArray, s[i...] == s.parent[map(-, i, s.shifts)...] if map(-, i, s.shifts) is a valid index for s.parent, and s.v[i, ...] == default otherwise. Use copy to collect the values of a ShiftedArray into a normal Array. The recommended constructor is ShiftedArray(parent, shifts; default = missing)

Examples

julia> v = [1, 3, 5, 4];

julia> s = ShiftedArray(v, (1,))
4-element ShiftedArrays.ShiftedArray{Int64,Missings.Missing,1,Array{Int64,1}}:
  missing
 1
 3
 5

julia> v = [1, 3, 5, 4];

julia> s = ShiftedArray(v, (1,))
4-element ShiftedArrays.ShiftedArray{Int64,Missings.Missing,1,Array{Int64,1}}:
  missing
 1
 3
 5

julia> copy(s)
4-element Array{Union{Int64, Missings.Missing},1}:
  missing
 1
 3
 5

julia> v = reshape(1:16, 4, 4);

julia> s = ShiftedArray(v, (0, 2))
4×4 ShiftedArrays.ShiftedArray{Int64,Missings.Missing,2,Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}}}:
 missing  missing  1  5
 missing  missing  2  6
 missing  missing  3  7
 missing  missing  4  8

julia> shifts(s)
(0, 2)

ShiftedArrays.ShiftedArrays

ShiftedArrays.ShiftedVector — Type.

ShiftedVector{T, S<:AbstractArray}

Shorthand for ShiftedArray{T, 1, S}.

ShiftedArrays._mapreduce

ShiftedArrays._mapreduce_vec

ShiftedArrays._padded_tuple

ShiftedArrays._shifted_between

ShiftedArrays.circshift — Function.

circshift(v::AbstractArray, n)

Return a CircShiftedArray object, with underlying data v. The second argument gives the amount to circularly shift in each dimension. If it is an integer, it is assumed to refer to the first dimension.

Examples

julia> v = [1, 3, 5, 4];

julia> ShiftedArrays.circshift(v, 1)
4-element ShiftedArrays.CircShiftedArray{Int64,1,Array{Int64,1}}:
 4
 1
 3
 5

julia> w = reshape(1:16, 4, 4);

julia> ShiftedArrays.circshift(w, (1, -1))
4×4 ShiftedArrays.CircShiftedArray{Int64,2,Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}}}:
 8  12  16  4
 5   9  13  1
 6  10  14  2
 7  11  15  3

ShiftedArrays.default — Function.

default(s::ShiftedArray)

Returns default value.

ShiftedArrays.eval

ShiftedArrays.get_circshifted_index

ShiftedArrays.include

ShiftedArrays.lag — Function.

lag(v::AbstractArray, n = 1; default = missing)

Return a ShiftedArray object, with underlying data v. The second argument gives the amount to shift in each dimension. If it is an integer, it is assumed to refer to the first dimension. default specifies a default value when you are out of bounds.

Examples

julia> v = [1, 3, 5, 4];

julia> lag(v)
4-element ShiftedArrays.ShiftedArray{Int64,Missings.Missing,1,Array{Int64,1}}:
  missing
 1
 3
 5

julia> w = 1:2:9
1:2:9

julia> s = lag(w, 2)
5-element ShiftedArrays.ShiftedArray{Int64,Missings.Missing,1,StepRange{Int64,Int64}}:
  missing
  missing
 1
 3
 5

julia> copy(s)
5-element Array{Union{Int64, Missings.Missing},1}:
  missing
  missing
 1
 3
 5

julia> v = reshape(1:16, 4, 4);

julia> s = lag(v, (0, 2))
4×4 ShiftedArrays.ShiftedArray{Int64,Missings.Missing,2,Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}}}:
 missing  missing  1  5
 missing  missing  2  6
 missing  missing  3  7
 missing  missing  4  8

ShiftedArrays.lead — Function.

lead(v::AbstractArray, n = 1; default = missing)

Return a ShiftedArray object, with underlying data v. The second argument gives the amount to shift negatively in each dimension. If it is an integer, it is assumed to refer to the first dimension. default specifies a default value when you are out of bounds.

Examples

julia> v = [1, 3, 5, 4];

julia> lead(v)
4-element ShiftedArrays.ShiftedArray{Int64,Missings.Missing,1,Array{Int64,1}}:
 3
 5
 4
  missing

julia> w = 1:2:9
1:2:9

julia> s = lead(w, 2)
5-element ShiftedArrays.ShiftedArray{Int64,Missings.Missing,1,StepRange{Int64,Int64}}:
 5
 7
 9
  missing
  missing

julia> copy(s)
5-element Array{Union{Int64, Missings.Missing},1}:
 5
 7
 9
  missing
  missing

julia> v = reshape(1:16, 4, 4);

julia> s = lead(v, (0, 2))
4×4 ShiftedArrays.ShiftedArray{Int64,Missings.Missing,2,Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}}}:
  9  13  missing  missing
 10  14  missing  missing
 11  15  missing  missing
 12  16  missing  missing

ShiftedArrays.mapreduce_vec — Function.

mapreduce_vec(g, f, ss::AbstractArray{<:ShiftedVector}, args...; default = missing, filter = t->true, dropmissing = true)

Align all vectors in ss. For each of the indices in args, extract the collection of elements corresponding to that index, apply g and reduce using f. f is any function that takes an iterable as input and outputs a scalar, such as mean. Indices for which the iterable is empty will return default.

Examples

julia> v = [1, 3, 5, 9, 6, 7];

julia> ss = ShiftedArray.((v,), [-2, -4])
2-element Array{ShiftedArrays.ShiftedArray{Int64,Missings.Missing,1,Array{Int64,1}},1}:
 Union{Int64, Missings.Missing}[5, 9, 6, 7, missing, missing]
 Union{Int64, Missings.Missing}[6, 7, missing, missing, missing, missing]

julia> mapreduce_vec(log, mean, ss, 1:2)
2-element Array{Float64,1}:
 1.7006
 2.07157

julia> mapreduce_vec(log, mean, ss, -5:2)
8-element Array{Any,1}:
  missing
  missing
 0.0
 1.09861
 0.804719
 1.64792
 1.7006
 2.07157

ShiftedArrays.reduce_vec — Function.

reduce_vec(f, ss::AbstractArray{<:ShiftedVector}, args...; default = missing, filter = t->true, dropmissing = true)

Align all vectors in ss. For each of the indices in args, extract the collection of elements corresponding to that index and reduce using f. f is any function that takes an iterable as input and outputs a scalar, such as mean. Indices for which the iterable is empty will return default.

Examples

julia> v = [1, 3, 5, 9, 6, 7];

julia> ss = ShiftedArray.((v,), [-2, -4])
2-element Array{ShiftedArrays.ShiftedArray{Int64,Missings.Missing,1,Array{Int64,1}},1}:
 Union{Int64, Missings.Missing}[5, 9, 6, 7, missing, missing]
 Union{Int64, Missings.Missing}[6, 7, missing, missing, missing, missing]

julia> reduce_vec(mean, ss, 1:2)
2-element Array{Float64,1}:
 5.5
 8.0

julia> reduce_vec(mean, ss, -5:2)
8-element Array{Any,1}:
  missing
  missing
 1.0
 3.0
 3.0
 6.0
 5.5
 8.0

ShiftedArrays.shifts — Function.

shifts(s::ShiftedArray)

Returns amount by which s is shifted compared to parent(s).

shifts(s::CircShiftedArray)

Returns amount by which s is shifted compared to parent(s).

ShiftedArrays.to_array — Function.

to_array(ss::AbstractArray{<:AbstractArray{<:Any, N}}, I::Vararg{<:AbstractArray, N}) where N

Collect a AbstractArray of AbstractArrays into a normal Array selecting indices I (can take negative values if inner AbstractArrays allow it). The output Array first few dimensions will be indexed by I (though starting from 1) and the last one will correspond to the index of the inner AbstractArray within the AbstractArray of AbstractArrays.

ShiftedArrays.to_offsetarray — Function.

to_offsetarray(ss::AbstractArray{<:AbstractArray{<:Any, N}}, I::Vararg{<:AbstractArray, N}) where N

Collect a AbstractArray of ShiftedArrays into an OffsetArray selecting indices I (can take negative values if inner AbstractArrays allow it). The output OffsetArray first few dimensions will be indexed by I and the last one will correspond to the index of the inner AbstractArray within the AbstractArray of AbstractArrays.