Docstrings

BlackBoxOptim.ADE_DefaultOptions

BlackBoxOptim.AbstractIndividual — Type.

The point of the SearchSpace.

The abstract type. It allows different actual implementations to be used, e.g Vector or SubArray.

BlackBoxOptim.AbstractPopulationMatrix

BlackBoxOptim.AdaptiveDiffEvoParameters — Type.

Specific data and functions for adaptation An Adaptive DE typically changes parameters of the search dynamically. This is typically done in the tell!() function when we know if the trial vector was better than the target vector.

BlackBoxOptim.AdaptiveDiffEvoRandBin — Type.

An Adaptive DE crossover operator changes cr and f parameters of the search dynamically.

BlackBoxOptim.AdaptiveDiffEvoRandBin1 — Type.

An Adaptive DE crossover operator changes cr and f parameters of the search dynamically.

BlackBoxOptim.AdaptiveDiffEvoRandBin2 — Type.

An Adaptive DE crossover operator changes cr and f parameters of the search dynamically.

BlackBoxOptim.Archive — Type.

Archive saves information about promising solutions during an optimization run.

BlackBoxOptim.ArchiveOutput — Type.

Base class for archive-specific component of the OptimizationResults.

BlackBoxOptim.ArchivedIndividual — Type.

Base class for individuals stored in Archive.

BlackBoxOptim.AskTellOptimizer — Type.

Base abstract class for optimizers that perform ask() → ..eval fitness.. → tell!() sequence at each step.

BlackBoxOptim.BimodalCauchy — Type.

A mixture of 2 Cauchy distributions. Random values are further constrained to [0.0, 1.0] range either by truncating the initial unconstrained value or by generating new random value until it fits the range.

BlackBoxOptim.BlackBoxOptim

BlackBoxOptim.BorgMOEA — Type.

Borg MOEA algorithm.

Based on Hadka & Reed, "Borg: An Auto-Adaptive Many-Objective Evolutionary Computing Framework", Evol. Comp. 2013

BlackBoxOptim.BorgMOEA_DefaultParameters

BlackBoxOptim.CEC09_UP8 — Function.

CEC09 Unconstrained Problem 8 objective function.

BlackBoxOptim.CEC09_UP8_PF — Function.

CEC09 Unconstrained Problem 8 Pareto Frontier. Parameterized by t[1]=f[1] and t[2]=f[2].

BlackBoxOptim.CEC09_Unconstrained_Set — Constant.

The collection of CEC09 unconstrained multi-objective problems.

See http://dces.essex.ac.uk/staff/zhang/MOEAcompetition/cec09testproblem0904.pdf.pdf

BlackBoxOptim.Candidate — Type.

Candidate solution to the problem.

Candidate can be either a member of the population (index > 0) or a standalone solution (index == -1). Can carry additional information, like the tag or the genetic operator applied (extra).

BlackBoxOptim.ChannelRef

BlackBoxOptim.ComplexFitnessScheme — Type.

Complex-valued fitness.

BlackBoxOptim.ConstantDirectionGen

BlackBoxOptim.ContinuousSearchSpace — Type.

Fixed-dimensional space, each dimension has a continuous range of valid values.

BlackBoxOptim.CrossoverOperator — Type.

Modifies NC "children" by transferring some information from NP "parents".

The concrete implementations must provide apply!() method.

BlackBoxOptim.DEX_DefaultOptions

BlackBoxOptim.DE_DefaultOptions

BlackBoxOptim.DXNESOpt — Type.

DX-NES: distance-weighted extensions of xNES by Fukushima et al.

BlackBoxOptim.DXNES_DefaultOptions

BlackBoxOptim.DefaultParameters — Constant.

Default parameters for all convenience methods that are exported to the end user.

See OptRunController for the description.

BlackBoxOptim.DictChain — Type.

An ordered set of dicts that are examined one after another to find the parameter value. Returns nothing if no param setting is found in any of the dicts.

BlackBoxOptim.DiffEvoCrossoverOperator

BlackBoxOptim.DiffEvoOpt

BlackBoxOptim.DiffEvoRandBin

BlackBoxOptim.DiffEvoRandBin1

BlackBoxOptim.DiffEvoRandBin2

BlackBoxOptim.DirectSearchProbabilisticDescentDefaultParameters

BlackBoxOptim.DirectSearcher — Type.

A supertype for all generating set searcher-like algorithms.

BlackBoxOptim.DirectionGenerator — Type.

DirectionGenerator generates the search directions to use at each step of a GSS search.

BlackBoxOptim.DummyMethodOutput

BlackBoxOptim.EMPTY_PARAMS — Constant.

The default placeholder value for parameters argument.

BlackBoxOptim.EmbeddingOperator — Type.

Embeds(projects) the individual into the search space.

The concrete implementations must provide apply!() method.

BlackBoxOptim.EpsBoxArchive — Type.

ϵ-box archive saves only the solutions that are not ϵ-box dominated by any other solutions in the archive.

It also counts the number of candidate solutions that have been added and how many ϵ-box progresses have been made.

BlackBoxOptim.EpsBoxArchiveFrontierIterator — Type.

Iterates occupied elements of the archive.frontier.

BlackBoxOptim.EpsBoxArchiveOutput — Type.

EpsBoxArchive-specific components of the optimization results.

BlackBoxOptim.EpsBoxArchive_DefaultParameters

BlackBoxOptim.EpsBoxDominanceFitnessScheme — Type.

EpsBoxDominanceFitnessScheme defines ϵ-box dominance for N-tuple (N≧1) fitnesses. It operates with fitnesses of type IndexedTupleFitness.

aggregator::AGG is a function mapping tuple fitness to a single numerical value. Might be used for comparisons (or not, depending on the setup). Always used when printing fitness vectors though.

BlackBoxOptim.EpsBoxFrontierIndividual — Type.

Individual stored in EpsBoxArchive.

BlackBoxOptim.EpsDominanceFitnessScheme — Type.

ϵ-dominance for N-tuple (N≧1) fitnesses.

aggregator::AGG is a function mapping tuple fitness to a single numerical value. Might be used for comparisons (or not, depending on the setup). Always used when printing fitness vectors though.

BlackBoxOptim.Evaluator — Type.

The abstract base for types that manage the objective function evaluation. P is the optimization problem it is used for.

BlackBoxOptim.ExponentialNaturalEvolutionStrategyOpt — Type.

Abstract type for a family of NES methods that represent population as

x = μ + σ B⋅Z,

where B is an exponential of some symmetric matrix lnB, tr(lnB)==0.0

BlackBoxOptim.FAGeneticOperatorsMixture — Type.

Frequency-adapting genetic operators mixture.

BlackBoxOptim.FitIndividual — Type.

A point in the problem's search space with the corresponding fitness value.

F is the original problem's fitness type

BlackBoxOptim.FitPopulation — Type.

The default implementation of PopulationWithFitness{F}.

BlackBoxOptim.FitnessScheme — Type.

FitnessScheme defines how fitness vectors/values are compared, presented and aggregated. Fitness represents the score of one and the same individual on one or a set of evaluations. A scheme is a specific way to consider the scores in a coherent way. Type parameter F specifies the type of fitness values.

FitnessScheme could also be used as a function that defines the fitness ordering, i.e. fs(x, y) == true iff fitness x is better than y.

BlackBoxOptim.FixedDimensionSearchSpace — Type.

SearchSpace with a fixed finite number of dimensions. Applicable to the vast majority of problems.

BlackBoxOptim.FixedGeneticOperatorsMixture — Type.

Randomly selects the genetic operator from the vector according to its weight and applies it.

BlackBoxOptim.FrequencyAdapter — Type.

A FrequencyAdapter adapts the frequencies with which a set of values/strategies/methods should be applied/tried in an optimization problem. It is based on the Adaptive Coordinate Frequencies scheme described in:

T. Glasmachers and U. Dogan, "Accelerated Coordinate Descent with Adaptive Coordinate Frequencies", 2013.

but generalized so that it can support more than the adaptation of only the coordinates in a Coordinate Descent scheme. The things that are being adapted are identified by integers in 1:n, with n being the main parameter.

BlackBoxOptim.FrontierIndividual — Type.

Individual representing the solution from the Pareto set.

BlackBoxOptim.FrontierIndividualWrapper — Type.

Wrapper for FrontierIndividual that allows easy access to the problem fitness.

BlackBoxOptim.FunctionBasedProblem — Type.

OptimizationProblem with the objective function defined by some Julia Function and search space.

Optionally, a known optimal value could be provided to terminate the optimization once it is reached.

BlackBoxOptim.FunctionBasedProblemFamily — Type.

Family of FunctionBasedProblem optimization problems parameterized by the number of search space dimensions.

BlackBoxOptim.GSSDefaultParameters

BlackBoxOptim.GeneratingSetSearcher — Type.

Generating Set Search as described in Kolda2003: Kolda, Tamara G., Robert Michael Lewis, and Virginia Torczon. "Optimization by direct search: New perspectives on some classical and modern methods." SIAM review 45.3 (2003): 385-482.

BlackBoxOptim.GeneticOperator — Type.

Abstract genetic operator that transforms individuals in the population.

BlackBoxOptim.GeneticOperatorsMixture — Type.

A mixture of genetic operators, use next() to choose the next operator from the mixture.

BlackBoxOptim.GibbsMutationOperator — Type.

Provides apply() operator that mutates one specified dimension of a parameter vector.

BlackBoxOptim.Hartman3

BlackBoxOptim.Hartman3_A

BlackBoxOptim.Hartman3_P

BlackBoxOptim.Hartman3_alpha

BlackBoxOptim.Hartman6

BlackBoxOptim.Hartman6_A

BlackBoxOptim.Hartman6_P

BlackBoxOptim.Hartman6_alpha

BlackBoxOptim.HatCompare

BlackBoxOptim.Hypersurface — Type.

$(N-1)$-dimensional manifold in $N$-dimensional space.

BlackBoxOptim.IGD — Function.

IGD(A::Vector{NTuple{N,F}}, B::Vector{NTuple{N,F}}, [two_sided=true])

The average Euclidean distance from the points of A to the points of B.

IGD(ref::Hypersurface, sol::Vector{FitIndividual}, [two_sided=true])

Average Euclidean distance from the exact Pareto frontier of the problem (ref) to the solution (sol) produced by the optimization method. If two_sided is on, returns the maximum of IGD(sol, ref) and IGD(nondominated(ref), sol).

BlackBoxOptim.IndexedTupleFitness — Type.

ϵ-box indexed representation of the N-tuple fitness.

Used together with EpsBoxDominanceFitnessScheme.

BlackBoxOptim.Individual — Type.

The point of the SearchSpace.

The concrete type that could be used for storage.

BlackBoxOptim.IndividualsSelector — Type.

Selects the individuals from the population.

The concrete implementations must provide select() method.

BlackBoxOptim.JadeFunctionSet — Constant.

JADE collection of optimization problems.

We skip (for now) f12 and f13 in the JADE paper since they are penalized functions which are quite nonstandard. We also skip f8 since we are unsure about its proper implementation.

BlackBoxOptim.MaximizingFitnessScheme

BlackBoxOptim.MethodNames — Constant.

Names of optimization methods accepted by bboptimize(), :Method keyword argument.

BlackBoxOptim.MethodOutput — Type.

Base class for method-specific component of the OptimizationResults.

BlackBoxOptim.MinimizationProblemFamily

BlackBoxOptim.MinimizingFitnessScheme

BlackBoxOptim.MirroredRandomDirectionGen — Type.

Generate half of the directions randomly and then mirrors by negating them.

BlackBoxOptim.MultiObjectiveMethodNames

BlackBoxOptim.MultiObjectiveMethods — Constant.

Multi-objective optimization methods accepted by bboptimize().

The values are the method initialization routines or types derived from Optimizer.

BlackBoxOptim.MutationClock — Type.

Mutation clock operator is a more efficient way to mutate vectors than to generate a random value per variable in the vectors. It instead generates the number of variables to skip until the next mutation. Then it uses a sub-mutation operator to do the actual mutation. This is based on the paper: Deb and Deb (2012), "Analyzing Mutation Schemes for Real-Parameter Genetic Algorithms" but we use a Poisson distribution.

BlackBoxOptim.MutationOperator — Type.

Modifies (mutates) one individual.

The concrete implementations must provide apply!() method.

BlackBoxOptim.MutationWrapper — Type.

Wraps the mutation operator, so that it could be used as crossover operator.

BlackBoxOptim.NES_DefaultOptions

BlackBoxOptim.NO_GEN_OP — Constant.

Placeholder for no-effect genetic operations.

BlackBoxOptim.NaturalEvolutionStrategyOpt

BlackBoxOptim.NoMutation — Type.

MutationOperator that does nothing.

BlackBoxOptim.OptController — Type.

Optimization Controller.

Applies specific optimization method to a given problem. Supports restarts and modifying parameter of the method between runs. runcontrollers field maintains the list of OptRunController instances applied so far.

See OptRunController.

BlackBoxOptim.OptRunController — Type.

Optimization Run Controller. Manages problem optimization using the specified method.

See OptController.

BlackBoxOptim.OptimizationProblem — Type.

The base abstract type for all optimization problems. FS is a type of a problem's FitnessScheme.

BlackBoxOptim.OptimizationResults — Type.

The results of running optimization method.

Returned by run!(oc::OptRunController). Should be compatible (on the API level) with the Optim package. See make_opt_results().

BlackBoxOptim.Optimizer — Type.

Base abstract class for black-box optimization algorithms.

BlackBoxOptim.PCX_DefaultOptions

BlackBoxOptim.PM_DefaultOptions — Constant.

Default parameters for PolynomialMutation.

BlackBoxOptim.ParallelEvaluationState — Type.

Current state of fitness function evaluation for the vector of candidates.

BlackBoxOptim.ParallelEvaluator — Type.

Fitness evaluator that distributes candidates fitness calculation among several worker processes.

BlackBoxOptim.ParallelEvaluatorWorker — Type.

Internal data for the worker process of the parallel evaluator.

BlackBoxOptim.ParallelEvaluatorWorkerRef

BlackBoxOptim.ParamBounds — Type.

The valid range of values for a specific dimension in a SearchSpace.

BlackBoxOptim.Parameters — Type.

The parameters storage type for BlackBoxOptim.

BlackBoxOptim.ParamsDict — Type.

The default parameters storage in BlackBoxOptim.

BlackBoxOptim.ParamsDictChain — Type.

An ordered set of dicts that are examined one after another to find the parameter value. Returns nothing if no param setting is found in any of the dicts.

BlackBoxOptim.ParentCentricCrossover — Type.

Parent Centric Crossover (PCX).

See Deb, K., Anand, A., and Joshi, D., "A Computationally Efficient Evolutionary Algorithm for Real-Parameter Optimization," Evolutionary Computation, vol. 10, no. 4, pp. 371-395, 2002.

BlackBoxOptim.ParetoFitnessScheme — Type.

Pareto dominance for N-tuple (N≧1) fitnesses.

aggregator::AGG is a function mapping tuple fitness to a single numerical value. Might be used for comparisons (or not, depending on the setup). Always used when printing fitness vectors though.

BlackBoxOptim.PolynomialMutation — Type.

Polynomial mutation as presented in the paper: Deb and Deb (2012), "Analyzing Mutation Schemes for Real-Parameter Genetic Algorithms"

BlackBoxOptim.Population — Type.

The base abstract type for the collection of candidate solutions in the population-based optimization methods.

BlackBoxOptim.PopulationMatrix — Type.

The simplest Population implementation – a matrix of floats, each column is an individual.

BlackBoxOptim.PopulationOptimizer — Type.

Base class for population-based optimization methods.

BlackBoxOptim.PopulationOptimizerOutput — Type.

PopulationOptimizer-specific components of the OptimizationResults. Stores the final population.

BlackBoxOptim.PopulationWithFitness — Type.

The base abstract types for population that also stores the candidates fitness.

F is the fitness type.

BlackBoxOptim.ProblemEvaluator — Type.

Default implementation of the Evaluator.

FP is the original problem's fitness type FA is the fitness type actually stored by the archive.

BlackBoxOptim.ProblemFamily — Type.

Base class for problem families.

It is an abstraction for problem parameterization (e.g by the number of the search space dimensions) that allows to instantiate OptimizationProblem for the concrete parameters.

BlackBoxOptim.Problems

BlackBoxOptim.RISDefaultParameters

BlackBoxOptim.RSDefaultParameters

BlackBoxOptim.RadiusLimitedSelector — Type.

IndividualsSelector that implements a "trivial geography" similar to Spector and Kline (2006) by first sampling an individual randomly and then selecting additional individuals for the same tournament within a certain deme of limited size (radius) for the sub-sequent individuals in the population.

The version we implement here is from: I. Harvey, "The Microbial Genetic Algorithm", in Advances in Artificial Life Darwin Meets von Neumann, Springer, 2011. The original paper is: Spector, L., and J. Klein. 2005. Trivial Geography in Genetic Programming. In Genetic Programming Theory and Practice III, edited by T. Yu, R.L. Riolo, and B. Worzel, pp. 109-124. Boston, MA: Kluwer Academic Publishers. http://faculty.hampshire.edu/lspector/pubs/trivial-geography-toappear.pdf

BlackBoxOptim.RandomBound — Type.

Embedding operator that randomly samples between parent's value and the nearest parameter boundary to get the new valid value if target's parameter is out-of-bounds.

BlackBoxOptim.RandomDirectionGen

BlackBoxOptim.RandomSearcher — Type.

Optimize by randomly generating the candidates.

BlackBoxOptim.RangePerDimSearchSpace — Type.

SearchSpace defined by a range of valid values for each dimension.

BlackBoxOptim.RatioFitnessScheme — Type.

In RatioFitnessScheme the fitness values can be ranked on a ratio scale so pairwise comparisons are not required. The default scale used is the aggregate of the fitness components.

BlackBoxOptim.ResamplingInheritanceMemeticSearcher

BlackBoxOptim.ResamplingMemeticSearcher — Type.

The variants of the memetic search algorithms RS and RIS. However, we have modified them since they did not give very good performance when implemented as described in the papers below. Possibly, the papers are not unambigous and I have misinterpreted something from them...

The "Resampling Search" (RS) memetic algorithm is described in:

F. Caraffini, F. Neri, M. Gongora and B. N. Passow, "Re-sampling Search: A
Seriously Simple Memetic Approach with a High Performance", 2013.

and its close sibling "Resampling Inheritance Search" (RIS) is described in:

F. Caraffini, F. Neri, B. N. Passow and G. Iacca, "Re-sampled Inheritance
Search: High Performance Despite the Simplicity", 2013.

BlackBoxOptim.SBX_DefaultOptions

BlackBoxOptim.SPSADefaultParameters

BlackBoxOptim.SPX_DefaultOptions

BlackBoxOptim.ScalarFitnessScheme — Type.

Float64-valued scalar fitness scheme. The boolean type parameter MIN specifies if smaller fitness values are better (true) or worse (false).

BlackBoxOptim.Schaffer1Family

BlackBoxOptim.SearchSpace — Type.

A SearchSpace defines the valid candidate points that could be considered in a search/optimization. The base abstract class has very few restrictions and can allow varying number of dimensions etc.

BlackBoxOptim.SeparableNESOpt — Type.

Separable Natural Evolution Strategy (sNES) optimizer.

BlackBoxOptim.Shekel10

BlackBoxOptim.Shekel10_A

BlackBoxOptim.Shekel10_C

BlackBoxOptim.Shekel5

BlackBoxOptim.Shekel5_A

BlackBoxOptim.Shekel5_C

BlackBoxOptim.Shekel7

BlackBoxOptim.Shekel7_A

BlackBoxOptim.Shekel7_C

BlackBoxOptim.ShiftedAndBiasedProblem — Type.

A TransformedProblem subclass that shifts the objective argument and offsets the value: $f(x - xshift) + fitshift$ → min/max, $x$ ∈ $X$ ($X$ stays intact).

BlackBoxOptim.SimpleSelector — Type.

Simple random IndividualsSelector.

The probabilties of all candidates are equal.

BlackBoxOptim.SimplexCrossover — Type.

Simplex Crossover (SPX).

ϵ>0 controls how the original simplex is inflated, ϵ=1 means no inflation.

See Tsutsui, Yamamura & Higuchi "Multi-parent recombination with simplex crossover in real coded genetic algorithms", 1999, Proc. of the Genetic and Evolutionary Computation Conference

BlackBoxOptim.SimulatedBinaryCrossover — Type.

Simulated Binary Crossover (SBX).

See Deb&Agrawal "Simulated binary crossover for continuous search space", 1994, Complex Systems

BlackBoxOptim.SimultaneousPerturbationSA2

BlackBoxOptim.SingleObjectiveMethodNames

BlackBoxOptim.SingleObjectiveMethods — Constant.

Single objective optimization methods accepted by bboptimize().

The values are the method initialization routines or types derived from Optimizer.

BlackBoxOptim.SteppingOptimizer — Type.

Optimizers derived from SteppingOptimizer implement classical iterative optimization scheme step!() → step!() → …

BlackBoxOptim.StochasticApproximationOptimizer — Type.

AskTellOptimizer that utilizes randomization to generate the candidates.

BlackBoxOptim.TopListArchive — Type.

Archive that maintains a top list of the best performing (best fitness) candidates seen so far.

The two last best fitness values could be used to estimate a confidence interval for how much improvement potential there is.

BlackBoxOptim.TopListArchiveOutput — Type.

TopListArchive-specific components of the optimization results.

BlackBoxOptim.TopListFitness — Type.

Fitness as stored in TopListArchive.

BlackBoxOptim.TopListIndividual — Type.

Individual stored in TopListArchive.

BlackBoxOptim.TournamentSelector — Type.

Tournament selector.

BlackBoxOptim.TransformedProblem — Type.

TransformedProblem{FS, P} is an abstract class for optimization problems derived from some original problem of type P by introducing a few changes.

The concrete derived types must implement:

objfunc() method

fitness_scheme() and search_space()

BlackBoxOptim.TupleFitnessScheme — Type.

Base class for tuple-based fitness schemes.

Type parameters:

N is the number of the objectives

F is the type of each objective

FA is the actual type of the multi-objective fitness

MIN if objectives should be minimized or maximized

AGG the type of aggregator

BlackBoxOptim.UNDX_DefaultOptions

BlackBoxOptim.UniformMutation — Type.

Uniform mutation of a parameter vector.

BlackBoxOptim.UnimodalNormalDistributionCrossover — Type.

Unimodal Normal Distribution Crossover (UNDX).

See Kita, H., Ono, I., and Kobayashi, S., "Multi-parental Extension of the Unimodal Normal Distribution Crossover for Real-coded Genetic Algorithms," Proceedings of the 1999 Congress on Evolutionary Computation, pp. 1581-1588, 1999. Deb, K., Anand, A., and Joshi, D., "A Computationally Efficient Evolutionary Algorithm for Real-Parameter Optimization," Evolutionary Computation, vol. 10, no. 4, pp. 371-395, 2002.

BlackBoxOptim.Utils

BlackBoxOptim.XNESOpt — Type.

xNES method.

Nice but scales badly with increasing dimensions.

BlackBoxOptim.XNES_DefaultOptions

BlackBoxOptim.ZERO_SEARCH_SPACE — Constant.

0-dimensional search space. Could be used as a placeholder for optional SearchSpace parameters.

BlackBoxOptim.abort!

BlackBoxOptim.accept_candi! — Function.

Put the candidate back into the pool and copy the values into the corresponding individual of the population (candi.index should be set).

BlackBoxOptim.ackley

BlackBoxOptim.acquire_candi — Function.

acquire_candi(pop::FitPopulation[, {ix::Int, candi::Candidate}])

Get individual from a pool, or create one if the pool is empty. By default the individual is not initialized, but if ix or candi is specified, the corresponding fields of the new candidate are set to the given values.

BlackBoxOptim.acquire_candis

BlackBoxOptim.acquire_mutant

BlackBoxOptim.adaptive_de_rand_1_bin

BlackBoxOptim.adaptive_de_rand_1_bin_radiuslimited

BlackBoxOptim.adaptive_diffevo

BlackBoxOptim.add_candidate! — Function.

add_candidate!(a::TopListArchive, fitness::F, candidate[, tag=0][, num_fevals=-1])

Add a candidate with a fitness to the archive (if it is good enough).

BlackBoxOptim.adjust! — Function.

Adjust the internal parameters of the genetic operator op taking into account the fitness change.

The default implementation does nothing.

Adjust the parameters of the method after the candidate evaluation.

BlackBoxOptim.aggregate — Function.

Aggregation is just the identity function for scalar fitness.

BlackBoxOptim.apply

BlackBoxOptim.apply! — Function.

Default implementation of apply!() for operators mixture.

BlackBoxOptim.archive

BlackBoxOptim.archived_fitness — Function.

archived_fitness(fit, a::Archive)

Converts given fitness fit to the format used by the archive a.

BlackBoxOptim.ask — Function.

ask(ato::AskTellOptimizer)

Ask for a new candidate solution to be generated, and a list of individuals it should be ranked with.

The individuals are supplied as an array of tuples with the individual and its index.

See also tell!()

BlackBoxOptim.assign_distance_weights!

BlackBoxOptim.assign_weights! — Function.

assign_weights!(weights, rankedCandidates, sortedWeights)

Assigns the candidate weights according to the candidate index. rankedCandidates are ranked by their fitness, sortedWeights are the corresponding weights.

Returns candidate weights sorted by the individual's index in the population.

BlackBoxOptim.bboptimize — Function.

bboptimize(problem[; parameters::Associative, kwargs...])

Solve given optimization problem.

See setup_problem() for the types of problem supported. In addition, the problem could be OptController containing the results of the previous optimization runs.

The optimization method parameters could be specified via kwargs or parameters arguments.

Returns OptimizationResults instance.

See also bbsetup().

BlackBoxOptim.bbsetup — Function.

bbsetup(problem[; parameters::Associative, kwargs...])

Set up optimization method for a given problem.

See setup_problem() for the types of problem supported. The optimization method parameters could be specified via kwargs or parameters arguments.

Returns the initialized OptController instance. To actually run the method call bboptimize() or run!().

BlackBoxOptim.best_candidate

BlackBoxOptim.best_fitness

BlackBoxOptim.best_of

BlackBoxOptim.borg_moea

BlackBoxOptim.calc_initial_step_size

BlackBoxOptim.calc_sigma_learnrate_for_nes

BlackBoxOptim.calc_sigma_learnrate_for_snes

BlackBoxOptim.calculate_evol_path_params — Function.

Calculate the parameters for evolutionary path.

Returns the tuple:

moving threshold

path discount

current Z scale

BlackBoxOptim.candi_pool_size

BlackBoxOptim.candidate_type

BlackBoxOptim.capacity

BlackBoxOptim.chain

BlackBoxOptim.check_and_create_search_space

BlackBoxOptim.check_epsbox_ϵ

BlackBoxOptim.check_stop_condition

BlackBoxOptim.check_valid!

BlackBoxOptim.cigar

BlackBoxOptim.cigtab

BlackBoxOptim.compare_optimizers

BlackBoxOptim.compass_search_directions

BlackBoxOptim.count_dict_report — Function.

Report the number of times each key was encountered in a count dict.

Returns a percentage dict calculated while iterating over the counted items.

BlackBoxOptim.crossover_parameters

BlackBoxOptim.de_rand_1_bin — Function.

The most used DE/rand/1/bin variant of differential evolution.

BlackBoxOptim.de_rand_1_bin_radiuslimited — Function.

The most used DE/rand/1/bin variant with "local geography" via radius-limited sampling.

BlackBoxOptim.de_rand_2_bin

BlackBoxOptim.de_rand_2_bin_radiuslimited

BlackBoxOptim.deceptive_cuccu2011 — Function.

Generator for the family of deceptive functions from the Cuccu2011 paper on novelty-based restarts. We have vectorized it to allow more than 1D versions. The Cuccu2011 paper uses the following values for

(l, w) = [(5, 0),  (15, 0),  (30, 0),
          (5, 2),  (15, 2),  (30, 2),
          (5, 10), (15, 10), (30, 10)]

and notes that (15, 2) and (30, 2) are the most difficult instances.

BlackBoxOptim.deceptive_cuccu2011_15_2

BlackBoxOptim.deceptive_cuccu2011_30_2

BlackBoxOptim.delta_fitness — Function.

delta_fitness(a::TopListArchive)

The difference between the current best fitness and the former best fitness.

BlackBoxOptim.deltas

BlackBoxOptim.describe

BlackBoxOptim.diameters

BlackBoxOptim.diffevo

BlackBoxOptim.direct_search_probabilistic_descent

BlackBoxOptim.directions_for_k

BlackBoxOptim.distance — Function.

distance(a::NTuple{N,F}, b::NTuple{N,F})

Euclidean distance from a to b.

BlackBoxOptim.distance_to_optimum — Function.

Get the distance from a fitness value to the optimum/best known fitness value.

BlackBoxOptim.dxnes

BlackBoxOptim.elapsed_time

BlackBoxOptim.ellipsoid — Function.

Schwefel's ellipsoid.

BlackBoxOptim.elliptic

BlackBoxOptim.energy_tsallis1996 — Function.

From section 3, page 7, of Tsallis1996 paper: Tsallis and Stariolo, "Generalized simulated annealing", Physica A, 1996. available from http://www.if.ufrgs.br/~stariolo/publications/TsSt96PhysA233395_1996.pdf the original paper used this as a 4-dimensional problem but here it is generalized.

BlackBoxOptim.eval

BlackBoxOptim.evaluator

BlackBoxOptim.evolve

BlackBoxOptim.evolved_pair — Function.

Post-process the evolved pair.

BlackBoxOptim.example_problems

BlackBoxOptim.f_calls

BlackBoxOptim.f_minimum

BlackBoxOptim.feasible — Function.

Projects a given point onto the search space coordinate-wise.

BlackBoxOptim.fill_block! — Function.

Fill the block of methods to apply.

BlackBoxOptim.fitness — Function.

fitness(x, p::OptimizationProblem)

Evaluates the fitness of a candidate.

fitness(indi::FitIndividual)

Gets the fitness of the individual.

fitness(params::Individual, e::ProblemEvaluator, tag::Int=0)

Evaluate the fitness and implicitly update the archive with the provided parameters and calculated fitness.

Returns the fitness in the archived format.

BlackBoxOptim.fitness_eltype

BlackBoxOptim.fitness_history_csv_header

BlackBoxOptim.fitness_improvement_potential — Function.

fitness_improvement_potential(a::Archive[, p = 0.01])

Calculate the solution improvement potential.

The percentage improvement that can be expected from the current fitness value at a given p-value. In theory, an optimization run should be terminated when this value is very small, i.e. there is little improvement potential left in the run.

BlackBoxOptim.fitness_improvement_ratio

BlackBoxOptim.fitness_is_within_ftol — Function.

Checks if the current best fitness is within the given range of a known best fitness. By default the best fitness is unknown.

BlackBoxOptim.fitness_scheme

BlackBoxOptim.fitness_scheme_type

BlackBoxOptim.fitness_shaping_utilities_linear — Function.

fitness_shaping_utilities_linear(n)

Calculate the n-dimensional fitness shaping utilities vector using the "steps" method.

BlackBoxOptim.fitness_shaping_utilities_log — Function.

fitness_shaping_utilities_log(n)

Calculate the n-dimensional fitness shaping utilities vector using the "log" method.

BlackBoxOptim.fitness_type — Function.

fitness_type(fs::FitnessScheme)
fitness_type(fs_type::Type{FitnessScheme})

Get the type of fitness values for fitness scheme fs.

BlackBoxOptim.fixed_dim_problem

BlackBoxOptim.flatten

BlackBoxOptim.format_fitness — Function.

format_fitness(fit, [problem::OptimizationProblem])

Format fitness into a string. Calls show_fitness() under the hood.

BlackBoxOptim.frequencies

BlackBoxOptim.general_stop_reason

BlackBoxOptim.generate — Function.

generate(surf::Hypersurface, fs::EpsBoxDominanceFitnessScheme, param_step = 0.1*fs.ϵ)

Generate the points of the hypersurface using the discretization defined by ϵ-box fitness schema.

BlackBoxOptim.griewank

BlackBoxOptim.hartman — Function.

Generic function for Hartman N problem family.

BlackBoxOptim.hartman3 — Function.

Hartman 3D is a multi-minima, non-separable test problem. Our implementation is based on: http://www.sfu.ca/~ssurjano/hart3.html However, we get a different global minima than the one stated on that page. The minima should be -3.86278 but we get a different one so use that in the problem spec:

BlackBoxOptim.hartman6 — Function.

Hartman 6D is a multi-minima, non-separable test problem. Our implementation is based on: http://www.sfu.ca/~ssurjano/hart6.html

BlackBoxOptim.has_converged

BlackBoxOptim.hat_compare — Function.

Check whether f1 or f2 fitness is better.

Returns

-1 if f1 is better than f2

1 if f2 is better than f1

0 if f1 and f2 are equal.

BlackBoxOptim.hat_compare_pareto

BlackBoxOptim.hat_compare_ϵ

BlackBoxOptim.hat_compare_ϵ_box — Function.

Returns a tuple of u and v comparison:

-1: u≺v

0: u and v non-dominating

1: u≻v

and whether u index fully matches v index.

BlackBoxOptim.include

BlackBoxOptim.index

BlackBoxOptim.ini_xnes_B — Function.

Identity for generic search space

Calculates the initial $log B$ matrix for xNES based on the deltas of each dimension.

BlackBoxOptim.init_rng!

BlackBoxOptim.instantiate — Function.

instantiate(family::FunctionBasedProblemFamily, ndim::Int)

Construct FunctionBasedProblem with the given number of dimensions.

BlackBoxOptim.instantiate_search_space — Function.

instantiate(family::FunctionBasedProblemFamily, ndim::Int)

Construct search space for FunctionBasedProblem with the given number of dimensions.

BlackBoxOptim.is_better

BlackBoxOptim.is_minimizing

BlackBoxOptim.is_stopped

BlackBoxOptim.is_worse

BlackBoxOptim.isinterrupted

BlackBoxOptim.isnafitness

BlackBoxOptim.isrunning

BlackBoxOptim.isstarted

BlackBoxOptim.isstopped

BlackBoxOptim.iteration_converged

BlackBoxOptim.iterations

BlackBoxOptim.kwargs2dict

BlackBoxOptim.last_fitness — Function.

last_fitness(e::Evaluator)

Get the fitness of the last evaluated candidate.

Leads to nicer code if we can first test if it is better or worse than existing candidates and only want the fitness itself if the criteria fulfilled.

BlackBoxOptim.last_top_fitness

BlackBoxOptim.lastrun

BlackBoxOptim.local_search

BlackBoxOptim.magnitude_class — Function.

Get a tuple of the sign and the magnitude of the value rounded to the first digit. Used for archiving candidates separately for each magnitude class.

BlackBoxOptim.make_evaluator — Function.

Create Evaluator instance for a given problem.

BlackBoxOptim.maxs

BlackBoxOptim.merge_fitness_histories — Function.

merge_fitness_histories(histories)

Merge the collection of multiple fitness histories and calculate the min, max, avg and median values for fitness and FIR (fitness improvement ratio) at each point where the fitness is changing.

BlackBoxOptim.minimization_problem

BlackBoxOptim.minimum

BlackBoxOptim.mins

BlackBoxOptim.nafitness

BlackBoxOptim.name

BlackBoxOptim.next — Function.

next(fa::FrequencyAdapter)

Give the index of the method that should be used at the next iteration.

FrequencyAdapter maintains a block of randomly shuffled methods. The function takes the next available method in the block. If the block is empty, it is repopulated.

The methods are randomly shuffled each time the block is regenerated, since we need to know their effectiveness at every period of the optimization process.

next(fa::FixedGeneticOperatorsMixture)

Gets the random genetic operator from the mixture.

The probability to select an operator is proportional to its weight.

Returns a tuple of the genetic operator and its index in the mix.

BlackBoxOptim.next_candidate! — Function.

Get the index of the next candidate for evaluation based on the Base.pmap() code.

BlackBoxOptim.noisy_quartic

BlackBoxOptim.nondominated — Function.

nondominated(fitnesses, fit_scheme)

Filter fitnesses removing all dominated values.

BlackBoxOptim.noprogress_streak — Function.

noprogress_streak(a::EpsBoxArchive, [since_restart])

Get the number of add_candidate!() calls since the last ϵ-progress. If since_restart is specified, the number is relative to the last restart.

BlackBoxOptim.notify!

BlackBoxOptim.num_evals

BlackBoxOptim.num_func_evals

BlackBoxOptim.num_steps

BlackBoxOptim.num_vars_to_next_mutation_point

BlackBoxOptim.numchildren

BlackBoxOptim.numdims

BlackBoxOptim.numobjectives

BlackBoxOptim.numparents

BlackBoxOptim.numruns

BlackBoxOptim.objfunc

BlackBoxOptim.occupied_frontier_indices

BlackBoxOptim.opt_value

BlackBoxOptim.optimizer

BlackBoxOptim.orig_problem

BlackBoxOptim.orig_problem_type

BlackBoxOptim.parameters

BlackBoxOptim.params — Function.

Get the problem parameters (a point in the search space) of the individual.

BlackBoxOptim.params_mean

BlackBoxOptim.params_std

BlackBoxOptim.pareto_frontier — Function.

Get the iterator to the individuals on the Pareto frontier.

BlackBoxOptim.persistent_individuals

BlackBoxOptim.persistent_range

BlackBoxOptim.popsize

BlackBoxOptim.populate_by_mutants

BlackBoxOptim.population — Function.

Generate a population for a given problem.

The default method to generate a population, uses Latin Hypercube Sampling.

BlackBoxOptim.problem

BlackBoxOptim.problem_set — Function.

problem_set(ps::Dict{Any, FunctionBasedProblemFamily}, dims::Union{Int,Vector{Int}})

Construct a fixed-dimensional version of each problem from ps for each dimension given in dims.

Returns a dictionary of problems.

BlackBoxOptim.problem_summary

BlackBoxOptim.process_candidate!

BlackBoxOptim.quartic

BlackBoxOptim.rand_frontier_index — Function.

Get random occupied Pareto frontier index. Returns 0 if frontier is empty.

BlackBoxOptim.rand_indexes

BlackBoxOptim.rand_individual — Function.

Generate one random candidate.

BlackBoxOptim.rand_individuals — Function.

Generate numIndividuals individuals by random sampling in the search space.

BlackBoxOptim.rand_individuals_lhs — Function.

Generate numIndividuals individuals by latin hypercube sampling (LHS). This should be the default way to create the initial population.

BlackBoxOptim.randbeta

BlackBoxOptim.random_resample

BlackBoxOptim.random_resample_with_inheritance

BlackBoxOptim.random_search

BlackBoxOptim.range_for_dim — Function.

Get the range of valid values for a specific dimension.

BlackBoxOptim.ranges

BlackBoxOptim.rank_by_fitness!

BlackBoxOptim.rank_result_dicts_by

BlackBoxOptim.rastrigin

BlackBoxOptim.recombinate!

BlackBoxOptim.release_candi — Function.

Put the candidate back to the pool.

BlackBoxOptim.repeated_bboptimize

BlackBoxOptim.report_from_result_dict — Function.

Print a report based on a result dict from one set of repeated runs of an optimization method. Returns the success rate, i.e. number of times the termination reason was "Within fitness tolerance...".

BlackBoxOptim.report_on_methods_results_on_one_problem

BlackBoxOptim.report_on_values — Function.

Summarize a vector of float values by stating its mean, std dev and median.

BlackBoxOptim.resampling_inheritance_memetic_searcher

BlackBoxOptim.resampling_memetic_searcher

BlackBoxOptim.reset! — Function.

Reset the current ParallelEvaluationState and the vector of candidates that need fitness evaluation.

BlackBoxOptim.reset_fitness! — Function.

Reset the candidate fitness.

Need it when the candidate parameters have changed, but the stored fitness is still for the old parameter set.

BlackBoxOptim.restart! — Function.

Restart Borg MOEA.

Resize and refills the population from the archive.

BlackBoxOptim.rosenbrock

BlackBoxOptim.run! — Function.

run!(ctrl::OptRunController)

Run optimization until one of the stopping conditions are satisfied.

run!(oc::OptController)

Start a new optimization run, possibly with new parameters and report on results.

BlackBoxOptim.s1_ackley

BlackBoxOptim.s1_elliptic

BlackBoxOptim.s1_rastrigin

BlackBoxOptim.s1_rosenbrock

BlackBoxOptim.s1_schwefel

BlackBoxOptim.s1_sphere

BlackBoxOptim.s2_step

BlackBoxOptim.same_fitness

BlackBoxOptim.sample_bernoulli_vector

BlackBoxOptim.sample_unit_hypersphere — Function.

Generate num random vectors on the n-dimensional, unit (hyper)sphere. This is the Muller-Marsaglia method as described here.

BlackBoxOptim.save_fitness_history_to_csv_file

BlackBoxOptim.schaffer1

BlackBoxOptim.schaffer1_PF

BlackBoxOptim.schwefel1_2

BlackBoxOptim.schwefel2_21

BlackBoxOptim.schwefel2_22

BlackBoxOptim.schwefel2_26

BlackBoxOptim.search_space

BlackBoxOptim.select — Function.

select(selector<:IndividualsSelector, population, numSamples::Int)

Select numSamples random candidates from the population.

BlackBoxOptim.separable_nes

BlackBoxOptim.set_as_elite_if_better

BlackBoxOptim.setup!

BlackBoxOptim.setup_optimizer!

BlackBoxOptim.setup_problem — Function.

setup_problem(problem, parameters::Parameters)

Set up a fixed-dimensional optimization problem.

There are several setup_problem() method that accept different type of problem argument: * OptimizationProblem * function (:NumDimensions has to be specified in parameters) * FunctionBasedProblemFamily (:NumDimensions has to be specified in parameters)

BlackBoxOptim.shekel — Function.

Generic function to define ShekelN problems.

BlackBoxOptim.shekel10 — Function.

Shekel10 is a 4D, multi-minima, non-separable test problem. Our implementation is based on the C code in: http://www.math.ntu.edu.tw/~wwang/colalab/testproblems/multipleopt/multioptprob/Shekel10/Shekel10.c

BlackBoxOptim.shekel5 — Function.

Shekel5 is a 4D, multi-minima, non-separable test problem. Our implementation is based on the C code in: http://www.math.ntu.edu.tw/~wwang/colalab/testproblems/multipleopt/multioptprob/Shekel5/Shekel5.c

BlackBoxOptim.shekel7 — Function.

Shekel7 is a 4D, multi-minima, non-separable test problem. Our implementation is based on the C code in: http://www.math.ntu.edu.tw/~wwang/colalab/testproblems/multipleopt/multioptprob/Shekel7/Shekel7.c

BlackBoxOptim.show_fitness — Function.

show_fitness(io, fit, [problem::OptimizationProblem])

Output fitness to the given IO stream. show_fitness() method could be overridden for a specific problem, e.g. to print the names of each objective.

BlackBoxOptim.show_report

BlackBoxOptim.shutdown!

BlackBoxOptim.shutdown_optimizer!

BlackBoxOptim.sphere

BlackBoxOptim.start_time

BlackBoxOptim.step

BlackBoxOptim.step! — Function.

step!(opt::SteppingOptimizer)

Do one iteration of the method.

BlackBoxOptim.stop_due_to_low_precision

BlackBoxOptim.stop_reason

BlackBoxOptim.symmetric_search_space — Function.

Create RangePerDimSearchSpace with given number of dimensions and given range of valid values for each dimension.

BlackBoxOptim.t_asy — Function.

Transform symmetric f into asymmetric objective function.

BlackBoxOptim.t_diag — Function.

Transform f into objective function with ill-conditioning effect.

BlackBoxOptim.t_irreg — Function.

Transform f into objective function with smooth local irregularities.

BlackBoxOptim.tag

BlackBoxOptim.tagcounts — Function.

tagcounts(a::EpsBoxArchive)

Count the tags of individuals on the ϵ-box frontier. Each restart the individual remains in the frontier discounts it by θ.

Returns the tag→count dictionary.

BlackBoxOptim.tell! — Function.

tell!(ato::AskTellOptimizer, rankedCandidates)

Tell the optimizer about the ranking of candidates. Returns the number of rankedCandidates that were inserted into the population, because of the improved fitness.

See also ask().

BlackBoxOptim.tournament — Function.

Simulate tournament among specified candidates.

Returns the index of the winner.

BlackBoxOptim.trace

BlackBoxOptim.trace_progress

BlackBoxOptim.trace_state — Function.

trace_state(io, op::GeneticOperator, mode::Symbol)

Trace the state of the operator. Called by trace_progress() during OptRunController run by some of the genetic optimizers.

Override the method to trace the state of your genetic operator.

trace_state(io::IO, optimizer::Optimizer, mode::Symbol)

Trace the current optimization state to a given IO stream. Called by OptRunController trace_progress().

Override it for your optimizer to generate method-specific diagnostic traces.

BlackBoxOptim.transient_range

BlackBoxOptim.update! — Function.

Update the internal model of progress and success rate of each method based on the latest progress value of one method. Progress values should be larger the larger the progress/improvement was.

BlackBoxOptim.update_candidates!

BlackBoxOptim.update_fitness!

BlackBoxOptim.update_learning_rates!

BlackBoxOptim.update_parameters! — Function.

update_parameters!(oc::OptController, parameters::Associative)

Update the OptController parameters.

BlackBoxOptim.update_population_fitness!

BlackBoxOptim.update_recombination_weights! — Function.

Update recombination operator probabilities based on the archive tag counts.

BlackBoxOptim.viewer — Function.

viewer(population, individual_index)

Get vector-slice of the population matrix for the specified individual. Does not allocate any additional space, while still providing the same lookup performance.

BlackBoxOptim.width_of_confidence_interval — Function.

Calculate the width of the confidence interval at a certain p-value. This is based on the paper: Carvalho (2011), "Confidence intervals for the minimum of a function using extreme value statistics"

This means that the current estimate of the confidence interval for the minimum of the optimized function lies within the interval

] l1 - w, l1 [

with probability $(1-p)$ as the number of sampled function points goes to infinity, where

w = width_of_confidence_interval(a, p)
l1 = best_fitness(a)

BlackBoxOptim.worker_finished! — Function.

Notify that the worker process is finished and reset its busy flag.

BlackBoxOptim.worst_fitness

BlackBoxOptim.write_result

BlackBoxOptim.xnes

BlackBoxOptim.xrotatedandshifted

BlackBoxOptim.xshifted

BlackBoxOptim.ϵ_index