Docstrings

Libz.@trans

Libz.@zcheck

Libz.DeflateSink

Libz.DeflateSource

Libz.InflateSink

Libz.InflateSource

Libz.Libz

Libz.Sink — Type.

The mode type parameter must be either :inflate or :deflate.

Libz.Source — Type.

The mode type parameter must be either :inflate or :deflate.

Libz.State

Libz.ZStream

Libz.Z_ASCII

Libz.Z_BEST_COMPRESSION

Libz.Z_BEST_SPEED

Libz.Z_BINARY

Libz.Z_BLOCK

Libz.Z_BUF_ERROR

Libz.Z_DATA_ERROR

Libz.Z_DEFAULT_COMPRESSION

Libz.Z_DEFAULT_STRATEGY

Libz.Z_DEFLATED

Libz.Z_ERRNO

Libz.Z_FILTERED

Libz.Z_FINISH

Libz.Z_FIXED

Libz.Z_FULL_FLUSH

Libz.Z_HUFFMAN_ONLY

Libz.Z_MEM_ERROR

Libz.Z_NEED_DICT

Libz.Z_NO_COMPRESSION

Libz.Z_NO_FLUSH

Libz.Z_OK

Libz.Z_PARTIAL_FLUSH

Libz.Z_RLE

Libz.Z_STREAM_END

Libz.Z_STREAM_ERROR

Libz.Z_SYNC_FLUSH

Libz.Z_TEXT

Libz.Z_TREES

Libz.Z_UNKNOWN

Libz.Z_VERSION_ERROR

Libz.ZlibDeflateInputStream — Function.

ZlibDeflateInputStream(input[; <keyword arguments>])

Construct a zlib deflate input stream to compress gzip/zlib data.

Arguments

input: a byte vector, IO object, or BufferedInputStream containing data to compress.

bufsize::Integer=8192: input and output buffer size.

raw::Bool=false: if true, data is raw compress data, without zlib metadata

gzip::Bool=true: if true, data is gzip compressed; if false, zlib compressed.

level::Integer=6: compression level in 1-9.

mem_level::Integer=8: memory to use for compression in 1-9.

strategy=Z_DEFAULT_STRATEGY: compression strategy; see zlib documentation.

Libz.ZlibDeflateOutputStream — Function.

ZlibDeflateOutputStream(output[; <keyword arguments>])

Construct a zlib deflate output stream to compress gzip/zlib data.

Arguments

output: a byte vector, IO object, or BufferedInputStream to which compressed data should be written.

bufsize::Integer=8192: input and output buffer size.

raw::Bool=false: if true, data is raw compress data, without zlib metadata

gzip::Bool=true: if true, data is gzip compressed; if false, zlib compressed.

level::Integer=6: compression level in 1-9.

mem_level::Integer=8: memory to use for compression in 1-9.

strategy=Z_DEFAULT_STRATEGY: compression strategy; see zlib documentation.

Libz.ZlibInflateInputStream — Function.

ZlibInflateInputStream(input[; <keyword arguments>])

Construct a zlib inflate input stream to decompress gzip/zlib data.

Arguments

input: a byte vector, IO object, or BufferedInputStream containing compressed data to inflate.

bufsize::Integer=8192: input and output buffer size.

raw::Bool=falso: if true, data is raw compress data, without zlib metadata

gzip::Bool=true: if true, data is gzip compressed; if false, zlib compressed.

reset_on_end::Bool=true: on stream end, try to find the start of another stream.

Libz.ZlibInflateOutputStream — Function.

ZlibInflateOutputStream(output[; <keyword arguments>])

Construct a zlib inflate output stream to decompress gzip/zlib data.

Arguments

output: a byte vector, IO object, or BufferedInputStream to which decompressed data should be written.

bufsize::Integer=8192: input and output buffer size.

raw::Bool=false: if true, data is raw compress data, without zlib metadata

gzip::Bool=true: if true, data is gzip compressed; if false, zlib compressed.

Libz.adler32 — Function.

adler32(data)

Compute the Adler-32 checksum over the data input. data can be BufferedInputStream or Vector{UInt8}.

Libz.code2str

Libz.compress

Libz.crc32 — Function.

crc32(data)

Compute the CRC-32 checksum over the data input. data can be BufferedInputStream or Vector{UInt8}.

Libz.decompress

Libz.deflate

Libz.deflate!

Libz.end_deflate!

Libz.end_inflate!

Libz.eval

Libz.finalized

Libz.finished

Libz.include

Libz.inflate

Libz.inflate!

Libz.init_deflate!

Libz.init_deflate_zstream

Libz.init_inflate!

Libz.init_inflate_zstream

Libz.initialized

Libz.inprogress

Libz.process

Libz.reset!

Libz.reset_deflate!

Libz.reset_inflate!

Libz.version

Libz.zerror

Libz.zlib

Libz.zlib_version