Docstrings

Graphics.@mustimplement

Graphics.BoundingBox — Type.

BoundingBox(xmin, xmax, ymin, ymax) -> bb

Create a representation bb of a rectangular region, specifying the coordinates of the horizontal (x) and vertical (y) edges.

Graphics.Graphics — Module.

Graphics defines an API for drawing in two dimensions.

Geometric primitives: Vec2, Point, BoundingBox

Geometry API: aspect_ratio, center, deform, diagonal, isinside, shift, height, width, xmin, xmax, ymin, ymax, xrange, yrange

2d drawing contexts: GraphicsDevice, GraphicsContext, creategc, getgc

Coordinate systems: set_coordinates, reset_transform, rotate, scale, translate, user_to_device!, device_to_user!, user_to_device_distance!, device_to_user_distance!, user_to_device, device_to_user

Lines: set_line_width, set_dash

Colors and painting (drawing attributes): set_source, set_source_rgb, set_source_rgba, save, restore

Clipping: clip, clip_preserve, reset_clip, inner_canvas

Paths: move_to, line_to, rel_line_to, rel_move_to, new_path, new_sub_path, close_path, arc

High-level paths: rectangle, circle, polygon

Fill and stroke: fill, fill_preserve, paint, stroke, stroke_preserve, stroke_transformed, stroke_transformed_preserve

Graphics.GraphicsContext

Graphics.GraphicsDevice

Graphics.Point — Type.

Point(x, y) -> p

Create a Cartesian representation p of a point in two dimensions.

Graphics.Vec2 — Type.

Vec2(x, y) -> v

Create a Cartesian representation v of a vector (or point) in two dimensions.

Graphics._set_coordinates

Graphics.arc

Graphics.aspect_ratio — Function.

aspect_ratio(bb::BoundingBox) -> r

Compute the ratio r of the height and width of bb.

Graphics.center

Graphics.circle

Graphics.clip

Graphics.clip_preserve

Graphics.close_path

Graphics.creategc

Graphics.d2ubuf

Graphics.deform — Function.

deform(bb::BoundingBox, Δl, Δr, Δt, Δb) -> bbnew

Add Δl (left), Δr (right), Δt (top), and Δb (bottom) to the edges of a BoundingBox.

Graphics.device_to_user

Graphics.device_to_user!

Graphics.device_to_user_distance!

Graphics.diagonal

Graphics.eval

Graphics.fill

Graphics.fill_preserve

Graphics.getgc

Graphics.height

Graphics.include

Graphics.inner_canvas — Function.

inner_canvas(c::GraphicsContext, device::BoundingBox, user::BoundingBox)
inner_canvas(c::GraphicsContext, x, y, w, h, l, r, t, b)

Create a rectangular drawing area inside device (represented in device-coordinates), giving it user-coordinates user. Any drawing that occurs outside this box is clipped.

x, y, w, and h are an alternative parametrization of device, and l, r, t, b parametrize user.

See also: set_coordinates.

Graphics.isinside — Function.

isinside(bb::BoundingBox, p::Point) -> tf::Bool
isinside(bb::BoundingBox, x, y) -> tf::Bool

Determine whether the point lies within bb.

Graphics.line_to

Graphics.move_to

Graphics.new_path

Graphics.new_sub_path

Graphics.paint

Graphics.polygon

Graphics.rectangle

Graphics.rel_line_to

Graphics.rel_move_to

Graphics.reset_clip

Graphics.reset_transform

Graphics.restore

Graphics.rotate — Function.

rotate(p::Vec2, angle::Real, o::Vec2) -> pnew

Rotate p around o by angle.

rotate(bb::BoundingBox, angle, o) -> bbnew

Rotate bb around o by angle, returning the BoundingBox that encloses the vertices of the rotated box.

Graphics.save

Graphics.scale

Graphics.set_coordinates — Function.

set_coordinates(c::GraphicsContext, device::BoundingBox, user::BoundingBox)
set_coordinates(c::GraphicsContext, user::BoundingBox)

Set the device->user coordinate transformation of c so that device, expressed in "device coordinates" (pixels), is equivalent to user as expressed in "user coordinates". If device is omitted, it defaults to the full span of c, BoundingBox(0, width(c), 0, height(c)).

See also get_matrix, set_matrix.

Graphics.set_coords

Graphics.set_dash

Graphics.set_line_width

Graphics.set_source

Graphics.set_source_rgb

Graphics.set_source_rgba

Graphics.shift — Function.

shift(bb::BoundingBox, Δx, Δy) -> bbnew

Shift center by (Δx,Δy), keeping width & height fixed.

Graphics.stroke

Graphics.stroke_preserve

Graphics.stroke_transformed

Graphics.stroke_transformed_preserve

Graphics.translate

Graphics.user_to_device

Graphics.user_to_device!

Graphics.user_to_device_distance!

Graphics.width

Graphics.with_aspect_ratio

Graphics.xmax

Graphics.xmin

Graphics.xrange

Graphics.ymax

Graphics.ymin

Graphics.yrange