Docstrings

DrakeVisualizer.ArrowHead

DrakeVisualizer.CommandQueue

DrakeVisualizer.Comms

DrakeVisualizer.CoreVisualizer

DrakeVisualizer.Cylinder

DrakeVisualizer.DrakeVisualizer

DrakeVisualizer.GeometryData

DrakeVisualizer.HyperCube

DrakeVisualizer.HyperEllipsoid

DrakeVisualizer.HyperRectangle

DrakeVisualizer.HyperSphere

DrakeVisualizer.LazyTrees

DrakeVisualizer.Link

DrakeVisualizer.MeshFile — Type.

A MeshFile is a wrapper around a mesh that indicates that the mesh should be transmitted to the viewer by saving its data to a file and sending that file's path to the viewer. This can avoid issues with very large meshes (which cannot otherwise be sent over LCM), but will not work if the viewer is not running locally.

DrakeVisualizer.Path

DrakeVisualizer.Point

DrakeVisualizer.PointCloud

DrakeVisualizer.PolyLine

DrakeVisualizer.Robot

DrakeVisualizer.Triad

DrakeVisualizer.Vec

DrakeVisualizer.VisData

DrakeVisualizer.VisTree

DrakeVisualizer.Visualizer

DrakeVisualizer.addgeometry!

DrakeVisualizer.any_open_windows

DrakeVisualizer.batch

DrakeVisualizer.center

DrakeVisualizer.contour_mesh

DrakeVisualizer.delete!

DrakeVisualizer.delete_director_binaries

DrakeVisualizer.drake_visualizer_executable_name

DrakeVisualizer.draw!

DrakeVisualizer.eval

DrakeVisualizer.include

DrakeVisualizer.intrinsic_transform

DrakeVisualizer.load!

DrakeVisualizer.new_window

DrakeVisualizer.onresponse

DrakeVisualizer.publish!

DrakeVisualizer.quaternion

DrakeVisualizer.radii

DrakeVisualizer.request_channel

DrakeVisualizer.response_channel

DrakeVisualizer.serialize

DrakeVisualizer.setgeometry!

DrakeVisualizer.settransform!

DrakeVisualizer.to_lcm

DrakeVisualizer.translation