Docstrings

RandomCorrelationMatrices.RandomCorrelationMatrices

RandomCorrelationMatrices.eval

RandomCorrelationMatrices.include

RandomCorrelationMatrices.randcormatrix — Function.

randcormatrix(d, η)

Generate a d-by-d random correlation matrix using the method described in:

Lewandowski, Kurowicka, Joe
  "Generating random correlation matrices based on vines
    and extended onion method"
Journal of Multivariate Analysis 100 (2009)
doi:10.1016/j.jmva.2009.04.008

RandomCorrelationMatrices.randcovmatrix — Function.

randcovmatrix(d, η, σ)

Use randcormatrix with desired vector of standard deviations σ for each term to generate a random covariance matrix.