Docstrings

OIFITS.Name

OIFITS.OIArray

OIFITS.OIContents

OIFITS.OICorrelation

OIFITS.OIData

OIFITS.OIDataBlock

OIFITS.OIDataBlockDef

OIFITS.OIFITS

OIFITS.OIFieldDef

OIFITS.OIFormatDef

OIFITS.OIMaster

OIFITS.OIPolarization

OIFITS.OISpectrum

OIFITS.OIT3

OIFITS.OITarget

OIFITS.OIVis

OIFITS.OIVis2

OIFITS.OIWavelength

OIFITS._BITPIX

OIFITS._COMMENT

OIFITS._DATABLOCKS

OIFITS._DATATYPE

OIFITS._DATATYPES

OIFITS._DTYPE_COMPLEX

OIFITS._DTYPE_INTEGER

OIFITS._DTYPE_LOGICAL

OIFITS._DTYPE_REAL

OIFITS._DTYPE_STRING

OIFITS._EXTENSION

OIFITS._EXTNAMES

OIFITS._FIELDS

OIFITS._FORMATS

OIFITS._REVERSE_DATATYPE

OIFITS.__init__

OIFITS.add_def

OIFITS.attach!

OIFITS.build_datablock

OIFITS.cfitsio_datatype

OIFITS.check_datablock

OIFITS.clone — Function.

Clone a data-block

This method clones an existing data-block. The array data are shared between the clones but not the links (owner, target, instrument, etc.). Only the fields defined by OI-FITS standard are cloned. This method is intended when a data-block from a given OIMaster instance is to be attached to another OIMaster instance.

OIFITS.default_revision

OIFITS.eval

OIFITS.fits_bitpix

OIFITS.fixname

OIFITS.get_amporder

OIFITS.get_amptyp

OIFITS.get_array — Function.

Assuming mst is an instance of OIMaster, then:

    get_array(mst, arrname)

yields the OIArray data-block of mst whose name is arrname if any, nothing otherwise.

Assuming db is an instance of a sub-type of OIDataBlock, then:

    get_array(db)

yields corresponding instance of OIArray if any, nothing otherwise.

OIFITS.get_arrays — Function.

Assuming mst is an instance of OIMaster, then:

    get_arrays(mst)

yields the names of the interferometric arrays defined in mst.

OIFITS.get_arrayx

OIFITS.get_arrayy

OIFITS.get_arrayz

OIFITS.get_arrname

OIFITS.get_calstat

OIFITS.get_category

OIFITS.get_comment

OIFITS.get_corr

OIFITS.get_corrindx_fluxdata

OIFITS.get_corrindx_ivis

OIFITS.get_corrindx_rvis

OIFITS.get_corrindx_t3amp

OIFITS.get_corrindx_t3phi

OIFITS.get_corrindx_vis2data

OIFITS.get_corrindx_visamp

OIFITS.get_corrindx_visphi

OIFITS.get_corrname

OIFITS.get_date_obs

OIFITS.get_dbname

OIFITS.get_dbtype

OIFITS.get_dec_err

OIFITS.get_decep0

OIFITS.get_def — Function.

get_def(db, revn = default_revision())

yields the OIDataBlockDef for datablock of type db (e.g., "OI_TARGET") in revison revn of OI-FITS standard.

OIFITS.get_descr

OIFITS.get_diameter

OIFITS.get_eff_band

OIFITS.get_eff_wave

OIFITS.get_equinox

OIFITS.get_file_handle

OIFITS.get_flag

OIFITS.get_fluxdata

OIFITS.get_fluxerr

OIFITS.get_fov

OIFITS.get_fovtype

OIFITS.get_frame

OIFITS.get_hdutype

OIFITS.get_iindx

OIFITS.get_insname

OIFITS.get_instrument — Function.

Assuming mst is an instance of OIMaster, then:

    get_instrument(mst, insname)

yields the OIWavelength data-block of mst whose name is insname if any, nothing otherwise.

Assuming db is an instance of a sub-type of OIDataBlock, then:

    get_instrument(db)

yields corresponding instance of OIWavelength if any, nothing otherwise.

OIFITS.get_instruments — Function.

Assuming mst is an instance of OIMaster, then:

    get_instruments(mst)

yields the names of the instruments defined in mst.

OIFITS.get_int_time

OIFITS.get_integer

OIFITS.get_ivis

OIFITS.get_iviserr

OIFITS.get_jindx

OIFITS.get_jxx

OIFITS.get_jxy

OIFITS.get_jyx

OIFITS.get_jyy

OIFITS.get_logical

OIFITS.get_mjd

OIFITS.get_mjd_end

OIFITS.get_mjd_obs

OIFITS.get_model

OIFITS.get_ndata

OIFITS.get_npol

OIFITS.get_orient

OIFITS.get_para_err

OIFITS.get_parallax

OIFITS.get_phiorder

OIFITS.get_phityp

OIFITS.get_pmdec

OIFITS.get_pmdec_err

OIFITS.get_pmra

OIFITS.get_pmra_err

OIFITS.get_ra_err

OIFITS.get_raep0

OIFITS.get_real

OIFITS.get_revn

OIFITS.get_rvis

OIFITS.get_rviserr

OIFITS.get_spectyp

OIFITS.get_sta_index

OIFITS.get_sta_name

OIFITS.get_staxyz

OIFITS.get_string

OIFITS.get_sysvel

OIFITS.get_t3amp

OIFITS.get_t3amperr

OIFITS.get_t3phi

OIFITS.get_t3phierr

OIFITS.get_target — Function.

Assuming mst is an instance of OIMaster, then:

    get_target(mst)

yields the OITarget data-block of mst if any, nothing otherwise.

Assuming tgt is an instance of OITarget, then:

    get_target(tgt)

yields the "TARGET" column of tgt which is an array of target names.

OIFITS.get_target_id — Function.

Assuming db is an instance of OITarget, OIVis, OIVis2 or OIT3, then:

    get_target_id(db)

yields the "TARGET_ID" column of db which is an array of integers.

OIFITS.get_targets — Function.

Assuming mst is an instance of OIMaster, then:

    get_targets(mst)

yields the names of the targets defined in mst.

OIFITS.get_tel_name

OIFITS.get_time

OIFITS.get_u1coord

OIFITS.get_u2coord

OIFITS.get_ucoord

OIFITS.get_v1coord

OIFITS.get_v2coord

OIFITS.get_value

OIFITS.get_vcoord

OIFITS.get_veldef

OIFITS.get_veltyp

OIFITS.get_vis2data

OIFITS.get_vis2err

OIFITS.get_visamp

OIFITS.get_visamperr

OIFITS.get_visphi

OIFITS.get_visphierr

OIFITS.get_visrefmap

OIFITS.hash_column_names

OIFITS.include

OIFITS.is_complex

OIFITS.is_integer

OIFITS.is_logical

OIFITS.is_real

OIFITS.is_string

OIFITS.load

OIFITS.make_hdu

OIFITS.name2Symbol

OIFITS.new_array

OIFITS.new_corr

OIFITS.new_inspol

OIFITS.new_master

OIFITS.new_spectrum

OIFITS.new_t3

OIFITS.new_target

OIFITS.new_vis

OIFITS.new_vis2

OIFITS.new_wavelength

OIFITS.read_column

OIFITS.read_datablock

OIFITS.read_table

OIFITS.readtable

OIFITS.select

OIFITS.select_target — Function.

Select data for a given target

The method select_target selects a subset of data corresponding to a given target. The general syntax is:

    out = select_target(inp, tgt)

where inp is the input data (can be an instance of OIMaster or of any OIDataBlock sub-types), tgt is the target number or name.

The result out may share part of its contents with the input data inp.

The result may be nothing if the input contains no data for the given target.

OIFITS.select_wavelength — Function.

Select data at given wavelengths

The method select_wavelength selects a subset of data on the basis of their wavelength. The general syntax is:

    out = select_wavelength(inp, sel)

where inp is the input data (can be an instance of OIMaster or of any OIDataBlock sub-types), sel is a predicate function which takes a wavelength (in meters) as argument and returns true if this wavelength is to be selected and false otherwise.

An alternative is:

    out = select_wavelength(inp, wmin, wmax)

to select wavelengths w such that wmin ≤ w ≤ wmax. The wavelength bounds are in meters.

The result out may share part of its contents with the input data inp.

The result may be nothing if the input contains no data at selected wavelengths.

OIFITS.symbolicname

OIFITS.to_complex

OIFITS.to_integer

OIFITS.to_real

OIFITS.update!