Docstrings

YaoBase.@assert_addr_inbounds

YaoBase.@assert_addrs

YaoBase.@assert_addrs_contiguous

YaoBase.@assert_addrs_inbounds

YaoBase.@interface

YaoBase.@lambda

YaoBase.@λ — Macro.

@λ <lambda definition>

Create legible lambdas.

Example

julia> @λ(x -> g(x)/3)
(x -> g(x)/3)

YaoBase.ASTTools — Module.

ASTTools

Tools for manipulating Julia ASTs. This module provides similar functionality with MacroTools, but based on MLStyle.

YaoBase.AbstractRegister — Type.

AbstractRegister{B, T}

Abstract type for quantum registers. B is the batch size, T is the data type.

YaoBase.AddressConflictError

YaoBase.AddressList

YaoBase.AddressNTuple

YaoBase.AddressVector

YaoBase.AdjointRegister — Type.

AdjointRegister{B, T, RT} <: AbstractRegister{B, T}

Lazy adjoint for a quantum register.

YaoBase.Const — Module.

Const

Contains common constant definitions.

YaoBase.LegibleLambda

YaoBase.NotImplementedError

YaoBase.QubitMismatchError — Type.

QubitMismatchError <: Exception

Qubit number mismatch error when applying a Block to a Register or concatenating Blocks.

YaoBase.YaoBase — Module.

Base module for Yao.

YaoBase._sort

YaoBase._wrap_identity

YaoBase.addbit!

BitBasis.basis — Function.

basis(register) -> UnitRange

Returns an UnitRange of the all the bits in the Hilbert space of given register.

source

YaoBase.batch_normalize — Function.

batch_normalize

normalize a batch of vector.

YaoBase.batch_normalize! — Function.

batch_normalize!(matrix)

normalize a batch of vector.

YaoBase.batched_kron

YaoBase.batched_kron!

Base.cat — Function.

cat(::AbstractRegister...) -> register

Merge several registers as one register via tensor product.

source

YaoBase.cutblock

YaoBase.cutlnn

YaoBase.datatype — Function.

datatype(register) -> Int

Returns the numerical data type used by register.

Note

datatype is not the same with eltype, since AbstractRegister family is not exactly the same with AbstractArray, it is an iterator of several registers.

source

YaoBase.define_abstract_api

YaoBase.density_fidelity — Function.

density_fidelity(ρ1, ρ2)

General fidelity (including mixed states) between two density matrix for qubits.

Definition

\[F(ρ, σ)^2 = tr(ρσ) + 2 \sqrt{det(ρ)det(σ)}\]

YaoBase.density_matrix — Function.

density_matrix(register)

Returns the density matrix of current active qubits.

source

YaoBase.eatwhere

YaoBase.eval

YaoBase.export_api

YaoBase.fidelity — Function.

fidelity(register1, register2)

Return the fidelity between two states.

Definition

The fidelity of two quantum state for qubits is defined as:

\[F(ρ, σ) = tr(\sqrt{\sqrt{ρ}σ\sqrt{ρ}})\]

Or its equivalent form (which we use in numerical calculation):

\[F(ρ, σ) = sqrt(tr(ρσ) + 2 \sqrt{det(ρ)det(σ)})\]

Reference

Jozsa R. Fidelity for mixed quantum states[J]. Journal of modern optics, 1994, 41(12): 2315-2323.

Nielsen M A, Chuang I. Quantum computation and quantum information[J]. 2002.

Note

The original definition of fidelity $F$ was from "transition probability", defined by Jozsa in 1994, it is the square of what we use here.

source

YaoBase.focus — Function.

focus(f, register, locs...)

Call a callable f under the context of focus. See also focus!.

Example

print the focused register

julia> r = ArrayReg(bit"101100")
ArrayReg{1,Complex{Float64},Array...}
    active qubits: 6/6

julia> focus(x->(println(x);x), r, 1, 2);
ArrayReg{1,Complex{Float64},Array...}
    active qubits: 2/6

source

YaoBase.focus! — Function.

focus!(register, locs) -> register

Focus the wires on specified location.

Example

julia> focus!(r, (1, 2, 4))

source

focus!(locs...) -> f(register) -> register

Lazy version of focus!, this returns a lambda which requires a register.

YaoBase.general_c1_gates — Function.

general_c1_gates(num_bit::Int, projector::AbstractMatrix, cbit::Int, gates::Vector{AbstractMatrix}, locs::Vector{Int}) -> AbstractMatrix

general (low performance) construction method for control gate on different lines.

YaoBase.general_controlled_gates — Function.

general_controlled_gates(num_bit::Int, projectors::Vector{Tp}, cbits::Vector{Int}, gates::Vector{AbstractMatrix}, locs::Vector{Int}) -> AbstractMatrix

Return general multi-controlled gates in hilbert space of num_bit qubits,

projectors are often chosen as P0 and P1 for inverse-Control and Control at specific position.

cbits should have the same length as projectors, specifing the controling positions.

gates are a list of controlled single qubit gates.

locs should have the same length as gates, specifing the gates positions.

YaoBase.get_raw_name

YaoBase.hilbertkron — Function.

hilbertkron(num_bit::Int, gates::Vector{AbstractMatrix}, locs::Vector{Int}) -> AbstractMatrix

Return general kronecher product form of gates in Hilbert space of num_bit qubits.

gates are a list of matrices.

start_locs should have the same length as gates, specifing the gates starting positions.

YaoBase.include

YaoBase.increase! — Function.

increase!(register, n::Int) -> register
increase!(n::Int) -> λ(register)

Increase the register by n bits in state |0>. i.e. |psi> -> |000> ⊗ |psi>, increased bits have higher indices.

If only an integer is provided, then returns a lambda function.

source

YaoBase.instruct! — Function.

instruct(state, operator[, locs, control_bits, control_vals])

instruction implementation for applying an operator to a quantum state.

This operator will be overloaded for different operator or state with different types.

YaoBase.invorder! — Function.

invorder(register)

Inverse the address of register.

source

YaoBase.is_function_call

YaoBase.isaddrs_conflict — Function.

isaddrs_conflict(addrs) -> Bool

Check if the input address has conflicts.

YaoBase.isaddrs_contiguous — Function.

isaddrs_contiguous(n::Int, addrs) -> Bool

Check if the input address is contiguous in $[1, n]$.

YaoBase.isaddrs_inbounds — Function.

isaddrs_inbounds(n, addrs) -> Bool

Check if the input address are inside given bounds n.

YaoBase.iscommute — Function.

iscommute(ops...) -> Bool

check if operators are commute.

LinearAlgebra.ishermitian — Function.

ishermitian(op) -> Bool

check if this operator is hermitian.

YaoBase.isreflexive — Function.

isreflexive(op) -> Bool

check if this operator is reflexive.

YaoBase.isunitary — Function.

isunitary(op) -> Bool

check if this operator is a unitary operator.

YaoBase.kron!

YaoBase.linop2dense — Function.

linop2dense([T=ComplexF64], linear_map!::Function, n::Int) -> Matrix

Returns the dense matrix representation given linear map function.

YaoBase.measure — Function.

measure(register[, locs]; ntimes=1) -> Vector{Int}

Return measurement results of current active qubits (regarding to active qubits, see focus! and relax!).

source

YaoBase.measure! — Function.

measure!(register[, locs])

Measure current active qubits or qubits at locs and collapse to result state.

source

YaoBase.measure_remove! — Function.

measure_remove!(::AbstractRegister[, locs])

Measure current active qubits or qubits at locs and remove them.

source

YaoBase.measure_reset!

YaoBase.measure_setto! — Function.

measure_setto!(reg::AbstractRegister[, locs]; bit_config) -> Int

Measure current active qubits or qubits at locs and set the register to specific value.

source

YaoBase.nactive — Function.

nactive(register) -> Int

Returns the number of active qubits.

Note

Operators always apply on active qubits.

source

YaoBase.nbatch — Function.

nbatch(register) -> Int

Returns the number of batches.

source

YaoBase.nonempty_maximum

YaoBase.nonempty_minimum

YaoBase.nqubits — Function.

nqubits(register) -> Int

Returns the (total) number of qubits. See nactive, nremain for more details.

source

YaoBase.nremain — Function.

nremain(register) -> Int

Returns the number of non-active qubits.

source

YaoBase.nucnorm — Function.

nucnorm(m)

Computes the nuclear norm of a matrix m.

YaoBase.probs — Function.

probs(register)

Returns the probability distribution of computation basis, aka $|<x|ψ>|^2$.

source

YaoBase.process_msgs

YaoBase.pure_state_fidelity — Function.

pure_state_fidelity(v1::Vector, v2::Vector)

fidelity for pure states.

YaoBase.purification_fidelity — Function.

purification_fidelity(m1::Matrix, m2::Matrix)

Fidelity for mixed states via purification.

Reference: http://iopscience.iop.org/article/10.1088/1367-2630/aa6a4b/meta

YaoBase.rand_hermitian — Function.

rand_hermitian(N::Int) -> Matrix

Random hermitian matrix.

YaoBase.rand_unitary — Function.

rand_unitary(N::Int) -> Matrix

Random unitary matrix.

YaoBase.relax! — Function.

relax!(register[, locs]; to_nactive=nqubits(register)) -> register

Inverse transformation of focus!, where to_nactive is the number of active bits for target register.

source

relax!(locs::Int...; to_nactive=nqubits(register)) -> f(register) -> register

Lazy version of relax!, it will be evaluated once you feed a register to its output lambda.

YaoBase.reorder! — Function.

reorder!(reigster, orders)

Reorder the address of register by input orders.

source

Base.repeat — Function.

repeat(r::AbstractRegister, n::Int) -> register

Repeat register r for n times on batch dimension.

Example

source

YaoBase.reset!

YaoBase.rotmat — Function.

rotmat(M::AbstractMatrix, θ::Real)

Returns rotated M: $exp(-\frac{imθ}{2} M)$.

YaoBase.select — Function.

select(register, bits...) -> AbstractRegister

Non-inplace version of select!.

source

YaoBase.select! — Function.

select!(dest::AbstractRegister, src::AbstractRegister, bits::Integer...) -> AbstractRegister
select!(register::AbstractRegister, bits::Integer...) -> register

select a subspace of given quantum state based on input eigen state bits. See also select.

Example

select!(reg, 0b110) will select the subspace with (focused) configuration 110. After selection, the focused qubit space is 0, so you may want call relax! manually.

Tip

Developers should overload select!(r::RegisterType, bits::NTuple{N, <:Integer}) and do not assume bits has specific number of bits (e.g Int64), or it will restrict the its maximum available number of qubits.

source

select!(b::Integer) -> f(register)

Lazy version of select!. See also select.

YaoBase.setto! — Function.

setto!(register, bit_str)

Set the register to bit string literal bit_str. About bit string literal, see more in @bit_str.

source

setto!(register, bit_config::Integer)

Set the register to bit configuration bit_config.

source

YaoBase.show

YaoBase.to_string

YaoBase.trace_distance — Function.

trace_distance(register1, register2)

Return the trace distance of register1 and register2.

Definition

Trace distance is defined as following:

\[\frac{1}{2} || A - B ||_{tr}\]

Reference

https://en.wikipedia.org/wiki/Trace_distance

source

YaoBase.trnorm — Function.

trnorm(m)

Computes the trace norm of a matrix m.

YaoBase.tupleargs

YaoBase.viewbatch — Function.

viewbatch(register, i::Int) -> AbstractRegister{1}

Returns a view of the i-th slice on batch dimension.

source

YaoBase.ρ — Function.

ρ(register)

Returns the density matrix of current active qubits. This is the same as density_matrix.

source