Docstrings

NaiveBayes.DataStats

NaiveBayes.GaussianNB

NaiveBayes.HybridNB — Type.

Initialize a HybridNB model with continuous and/or discrete features

Constructors

HybridNB(labels::AbstractVector, kde_names::AbstractVector, discrete_names::AbstractVector)
HybridNB(labels::AbstractVector, kde_names::AbstractVector)
HybridNB(labels::AbstractVector, num_kde::Int, num_discrete::Int)

Arguments

labels : An AbstractVector{Any} of feature labels

kde_names : An AbstractVector{Any} of the names of continuous features

discrete_names : An AbstractVector{Any} of the names of discrete features

num_kde : Number of continuous features

num_discrete : Number of discrete features

NaiveBayes.KernelNB

NaiveBayes.MultinomialNB — Type.

Multinomial Naive Bayes classifier

classes : array of objects Class names n_vars : Int64 Number of variables in observations alpha : Number (optional, default 1) Smoothing parameter. E.g. if alpha equals 1, each variable in each class is believed to have 1 observation by default

NaiveBayes.NBModel — Type.

Base type for Naive Bayes models. Inherited classes should have at least following fields: ccounts::Dict{C, Int64} - count of ocurrences of each class nobs::Int64 - total number of observations

NaiveBayes.NaiveBayes

NaiveBayes.cov

NaiveBayes.ePDF — Type.

a wrapper around key value pairs for a discrete probability distribution

NaiveBayes.ensure_data_size

NaiveBayes.eval

StatsBase.fit — Function.

fit(m::HybridNB, f_c::Vector{Vector{Float64}}, f_d::Vector{Vector{Int64}}, labels::Vector{Int64})

Train NB model with discrete and continuous features by estimating P(x⃗|c)

fit(m::HybridNB, f_c::Matrix{Float64}, labels::Vector{Int64})

Train NB model with continuous features only

NaiveBayes.get_feature_names

NaiveBayes.include

NaiveBayes.load_model

NaiveBayes.logprob_c

NaiveBayes.logprob_x_given_c — Function.

Calculate log P(x|C)

Calculate log P(x|C)

Calculate log P(x|C)

Calculate log P(x|C)

NaiveBayes.mean

NaiveBayes.num_discrete

NaiveBayes.num_features

NaiveBayes.num_kdes

NaiveBayes.num_samples — Function.

compute the number of samples

StatsBase.predict — Function.

Predict kde naive bayes for continuos featuers only

predict(m::HybridNB, f_c::Vector{Vector{Float64}}, f_d::Vector{Vector{Int64}}) -> labels

Predict hybrid naive bayes for continuos featuers only

NaiveBayes.predict_logprobs — Function.

Predict log probabilities for all classes

Predict log probabilities for all classes

predict_logprobs(m::HybridNB, features_c::Vector{Vector{Float64}, features_d::Vector{Vector{Int})

Return the log-probabilities for each column of X, where each row is the class

NaiveBayes.predict_proba — Function.

Predict logprobs, return tuples of predicted class and its logprob

predict_proba{V<:Number}(m::HybridNB, f_c::Vector{Vector{Float64}}, f_d::Vector{Vector{Int64}})

Predict log-probabilities for the input features. Returns tuples of predicted class and its log-probability estimate.

NaiveBayes.probability — Function.

query the ePDF to get the probability of n

NaiveBayes.restructure_matrix — Function.

restructure_matrix(M::Matrix) -> V::Dict{Symbol, Vector}

Restructure a matrix as vector of vectors

NaiveBayes.sum_log_x_given_c! — Function.

computes log[P(x⃗ⁿ|c)] ≈ ∑ᵢ log[p(xⁿᵢ|c)]

NaiveBayes.to_matrix — Function.

to_matrix(D::Dict{Symbol, Vector}}) -> M::Matrix

convert a dictionary of vectors into a matrix

NaiveBayes.to_range

NaiveBayes.train — Function.

train(HybridNB, continuous, discrete, labels) -> model2

NaiveBayes.updatestats

NaiveBayes.write_model