Docstrings

Electron.Application — Type.

function Application()

Start a new Electron application. This will start a new process for that Electron app and return an instance of Application that can be used in the construction of Electron windows.

Electron.Electron

Electron.JSError

Electron.OptDict

Electron.URI

Electron.URI_file — Function.

URI_file(base, filespec)

Construct an absolute URI to filespec relative to base.

Electron.Window — Type.

function Window([app::Application,] options::Dict)

Open a new Window in the application app. Pass the content of options to the Electron BrowserWindow constructor.

If app is not specified, use the default Electron application, starting one if needed.

function Window([app::Application,] uri::URI)

Open a new Window in the application app. Show the content that uri points to in that new window.

If app is not specified, use the default Electron application, starting one if needed.

function Window([app::Application,] uri::URI)

Open a new Window in the application app. Show the content as a text/html file with utf-8 encoding.

If app is not specified, use the default Electron application, starting one if needed.

Electron._Application

Electron._Window

Electron.__init__

Electron._global_applications

Electron.applications

Electron.default_application

Electron.eval

Electron.generate_pipe_name

Electron.get_electron_binary_cmd

Electron.include

Electron.load — Function.

load(win::Window, uri::URI)

Load uri in the Electron window win.

load(win::Window, html::AbstractString)

Load html in the Electron window win.

Electron.msgchannel

Electron.req_response

Electron.toggle_devtools

Electron.windows