Docstrings

Yeppp.@checked_lib

Yeppp.@yepppfunsAA_A

Yeppp.@yepppfunsA_A

Yeppp.@yepppfunsA_S

Yeppp.Yeppp

Yeppp.__init__

Yeppp.add — Function.

add(x::Array, y::Array)

Perform element wise addition of the two array x and y.

Yeppp.add!

Yeppp.cos — Function.

cos(x)

Returns element wise cos of x.

Yeppp.cos! — Function.

cos(x)

Returns element wise cos of x.

Yeppp.dot — Function.

dot(x::Vector{Float64}, y::Vector{Float64})

Compute the dot product of x and y.

dot(x::Vector{Float32}, y::Vector{Float32})

Compute the dot product of x and y.

Yeppp.eval

Yeppp.evalpoly — Function.

evalpoly(coeff, x)

Evaluates polynomial with double precision (64-bit) floating-point coefficients coeff on an array of double precision (64-bit) floating-point elements x.

Yeppp.evalpoly!

Yeppp.exp — Function.

exp(x)

Returns the element wise exponential of x.

Yeppp.exp! — Function.

exp(x)

Computes the element wise exponential of x inplace.

Yeppp.include

Yeppp.libyeppp

Yeppp.log — Function.

log(x)

Returns the element wise natural logarithm of x.

Yeppp.log! — Function.

log!(x)

Computes the element wise natural logarithm of x inplace.

Yeppp.max — Function.

max(x, y)

Compares the vectors x and y and return the element wise maxima.

Yeppp.max!

Yeppp.min — Function.

max(x, y)

Compares the vectors x and y and return the element wise minima.

Yeppp.min!

Yeppp.multiply — Function.

multiply(x, y)

Perform element wise multiplication of the two array x and y.

Yeppp.multiply!

Yeppp.release

Yeppp.sin — Function.

sin(x)

Returns element wise sin of x.

Yeppp.sin! — Function.

sin(x)

Computes element wise sin of x inplace.

Yeppp.subtract — Function.

subtract(x::Array, y::Array)

Perform element wise subtraction of the two array x and y.

Yeppp.subtract!

Yeppp.sum

Yeppp.sumabs

Yeppp.sumabs2

Yeppp.tan — Function.

tan(x)

Returns element wise tan of x.

Yeppp.tan! — Function.

tan(x)

Returns element wise tan of x.