Docstrings

AffineInvariantMCMC.AffineInvariantMCMC — Module.

MADS: Model Analysis & Decision Support in Julia (Mads.jl v1.0) 2016 module AffineInvariantMCMC

https://mads.lanl.gov https://github.com/madsjulia

Licensing: GPLv3: http://www.gnu.org/licenses/gpl-3.0.html

Copyright 2016. Los Alamos National Security, LLC. All rights reserved.

This material was produced under U.S. Government contract DE-AC52-06NA25396 for Los Alamos National Laboratory, which is operated by Los Alamos National Security, LLC for the U.S. Department of Energy. The Government is granted for itself and others acting on its behalf a paid-up, nonexclusive, irrevocable worldwide license in this material to reproduce, prepare derivative works, and perform publicly and display publicly. Beginning five (5) years after –––––––- November 17, 2015, –––––––––––––––––––––––––––––––– subject to additional five-year worldwide renewals, the Government is granted for itself and others acting on its behalf a paid-up, nonexclusive, irrevocable worldwide license in this material to reproduce, prepare derivative works, distribute copies to the public, perform publicly and display publicly, and to permit others to do so.

NEITHER THE UNITED STATES NOR THE UNITED STATES DEPARTMENT OF ENERGY, NOR LOS ALAMOS NATIONAL SECURITY, LLC, NOR ANY OF THEIR EMPLOYEES, MAKES ANY WARRANTY, EXPRESS OR IMPLIED, OR ASSUMES ANY LEGAL LIABILITY OR RESPONSIBILITY FOR THE ACCURACY, COMPLETENESS, OR USEFULNESS OF ANY INFORMATION, APPARATUS, PRODUCT, OR PROCESS DISCLOSED, OR REPRESENTS THAT ITS USE WOULD NOT INFRINGE PRIVATELY OWNED RIGHTS.

LA-CC-15-080; Copyright Number Assigned: C16008

AffineInvariantMCMC.eval

AffineInvariantMCMC.flattenmcmcarray — Function.

Flatten MCMC arrays

AffineInvariantMCMC.include

AffineInvariantMCMC.sample — Function.

Bayesian sampling using Goodman & Weare's Affine Invariant Markov chain Monte Carlo (MCMC) Ensemble sampler (aka Emcee)

AffineInvariantMCMC.sample(llhood, numwalkers=10, numsamples_perwalker=100, thinning=1)

Arguments:

llhood : function estimating loglikelihood (for example, generated using Mads.makearrayloglikelihood())

numwalkers : number of walkers

x0 : normalized initial parameters (matrix of size (length(params), numwalkers))

thinning : removal of any thinning realization

a :

Returns:

mcmcchain : final MCMC chain

llhoodvals : log likelihoods of the final samples in the chain

Reference:

Goodman & Weare, "Ensemble samplers with affine invariance", Communications in Applied Mathematics and Computational Science, DOI: 10.2140/camcos.2010.5.65, 2010.