Docstrings

LSODA.@checked_lib

LSODA.@pack_lsoda_common_t

LSODA.@pack_lsoda_common_t!

LSODA.@pack_lsoda_context_t

LSODA.@pack_lsoda_context_t!

LSODA.@pack_lsoda_opt_t

LSODA.@pack_lsoda_opt_t!

LSODA.@unpack_lsoda_common_t

LSODA.@unpack_lsoda_context_t

LSODA.@unpack_lsoda_opt_t

LSODA.AbstractLSODAHandle

LSODA.CommonFunction

LSODA.ContextHandle

LSODA.LSODA

LSODA.LSODAAlgorithm

LSODA.UserFunctionAndData

LSODA.__init__

LSODA._lsoda_f

LSODA.commonfun

LSODA.depsfile

LSODA.eval

LSODA.include

LSODA.liblsoda

LSODA.lsoda — Type.

lsoda(f::Function, y0::Vector{Float64}, tspan::Vector{Float64}; userdata::Any=nothing, reltol::Union{Float64,Vector}=1e-4, abstol::Union{Float64,Vector}=1e-10)

Solves a set of ordinary differential equations using the LSODA algorithm. The vector field encoded in an inplace f::Function needs to have the self-explanatory arguments f(t, y, ydot, data)

LSODA.lsoda_0

LSODA.lsoda_common_t

LSODA.lsoda_context_t

LSODA.lsoda_context_t_ptr

LSODA.lsoda_evolve! — Function.

lsodaevolve!(ctx::lsodacontext_t,y::Vector{Float64},tspan::Vector{Float64})

Solves a set of ordinary differential equations using the LSODA algorithm and the context variable ctx. This avoid re-allocating ctx. You have to be carefull to remember the current time or this function will return an error.

LSODA.lsoda_free

LSODA.lsoda_opt_t

LSODA.lsoda_prepare

LSODA.lsoda_reset

LSODA.lsodafun

LSODA.old_cfunction

LSODA.release_handle

LSODA.solve

LSODA.warnkeywords

LSODA.warnlist