Docstrings

WordTokenizers.MERGESYMBOL — Constant.

A simple reversible tokenizer

tokenized = rev_tokenize(instring)
de_tokenized = rev_detokenize(token)

The revtokenize tokenizer splits into token based on space, punctuations and special symbols and in addition it leaves some merge-symbols ('') for the tokens to be re-arranged when needed using the revdetokenize. It uses a character based approach for splitting and re-merging.

Parameters:

instring: Input string to be tokenized

token: Collection to tokens i.e String Array

WordTokenizers.Sentences

WordTokenizers.TokenBuffer — Type.

TokenBuffer("foo bar")

Turns a string into a readable stream, used for building tokenisers. Utility parsers such as spaces and number read characters from the stream and into an array of tokens.

Parsers return true or false to indicate whether they matched anything in the input stream. They can therefore be combined easily, e.g.

spacesornumber(ts) = spaces(ts) || number(ts)

either skips whitespace or parses a number token, if possible.

The simplest possible tokeniser accepts any character with no token breaks:

function tokenise(input)
  ts = TokenBuffer(input)
  while !isdone(ts)
    character(ts)
  end
  return ts.tokens
end

tokenise("foo bar baz") # ["foo bar baz"]

The second simplest splits only on spaces:

function tokenise(input)
  ts = TokenBuffer(input)
  while !isdone(ts)
    spaces(ts) || character(ts)
  end
  return ts.tokens
end

tokenise("foo bar baz") # ["foo", "bar", "baz"]

See nltk_word_tokenize for a more advanced example.

WordTokenizers.WordTokenizers

WordTokenizers.Words

WordTokenizers.atoms — Function.

atoms(::TokenBuffer, ["--", "...", ...])

Matches a set of atomic tokens, such as ..., which should always be treated as a single token, regardless of word boundaries.

WordTokenizers.character — Function.

character(::TokenBuffer)

Push the next character in the input into the buffer's current token.

WordTokenizers.closingquote — Function.

openquote(::TokenBuffer)

Matches " used as a closing quote, and tokenises it as ''.

WordTokenizers.eval

WordTokenizers.flush! — Function.

flush!(::TokenBuffer, tokens...)

TokenBuffer builds the current token as characters are read from the input. When the end of the current token is detected, call flush! to finish it and append it to the token stream. Optionally, give additional tokens to be added to the stream after the current one.

WordTokenizers.generate_tokenizer_from_sed — Function.

generate_tokenizer_from_sed(sed_script_path)

This returns Julia code, that is the translation of a simple sed script for tokenizing. This doesn't fully cover all the functionality of sed, but it covers enough for many purposes.

WordTokenizers.improved_penn_tokenize — Function.

improved_penn_tokenize(input::AbstractString)

The Improved Penn Treebank tokenizer. This is a port of NLTK's modified Penn Tokenizer. It has a bundle of minor changes, that I don't think are actually documented anywhere. But things like cannot cannot become can not can not where as the original would produce can not cannot.

The tokenizer still seperates out contractions: "shouldn't" becomes ["should", "n't"]

The input should be a single sentence; but again it will likely be relatively fine if it isn't. Depends exactly what you want it for.

This matches NLTK's nltk.tokenize.TreeBankWordTokenizer.tokenize

WordTokenizers.include

WordTokenizers.is_weird

WordTokenizers.isdone

WordTokenizers.lookahead — Function.

lookahead(::TokenBuffer, s; boundary = false)

Peek at the input to see if s is coming up next. boundary specifies whether a word boundary should follow s.

julia> lookahead(TokenBuffer("foo bar"), "foo")
true

julia> lookahead(TokenBuffer("foo bar"), "bar")
false

julia> lookahead(TokenBuffer("foo bar"), "foo", boundary = true)
true

julia> lookahead(TokenBuffer("foobar"), "foo", boundary = true)
false

WordTokenizers.nltk_atoms

WordTokenizers.nltk_splits

WordTokenizers.nltk_suffixes

WordTokenizers.nltk_word_tokenize — Function.

nltk_word_tokenize(input)

NLTK's word tokenizer. It is an extension on the Punctuation Preserving Penn Treebank tokenizer, mostly to better handle unicode.

Punctuation is still preserved as its own token. This includes periods which will be stripped from words.

The tokenizer still seperates out contractions: "shouldn't" becomes ["should", "n't"]

The input should be a single sentence; but again it will likely be relatively fine if it isn't. Depends exactly what you want it for.

This matches to the most commonly used nltk.word_tokenize, minus the sentence tokenizing step.

WordTokenizers.nth_ind

WordTokenizers.number — Function.

number(::TokenBuffer)

Matches numbers such as 10,000.5, preserving formatting.

WordTokenizers.openquote — Function.

openquote(::TokenBuffer)

Matches " used as an opening quote, and tokenises it as ``.

WordTokenizers.penn_tokenize — Function.

penn_tokenize(input::AbstractString)

"... to produce Penn Treebank tokenization on arbitrary raw text. Yeah, sure" quote Robert MacIntyre

Tokenization does a number of things like seperate out contractions: "shouldn't" becomes ["should", "n't"] Most other punctuation becomes &'s. Exception is periods which are not touched. The input should be a single sentence; but it will likely be relatively fine if it isn't. Depends exactly what you want it for.

This is a direct (automatic) translation of the original sed script.

If you want to mess with exactly what it does it is actually really easy. copy the penn.sed file from this repo, modify it to your hearts content. There are some lines you can uncomment out. You can generate a new tokenizer using:

@generated function custom_tokenizer(input::AbstractString)
    generate_tokenizer_from_sed(joinpath(@__DIR__, "custom.sed"))
end

WordTokenizers.poormans_tokenize — Function.

poormans_tokenize

Tokenizes by removing punctuation and splitting on spaces

WordTokenizers.postproc_splits — Function.

 postproc_splits(sentences)

Applies heuristic rules to repair sentence splitting errors. Developed for use as postprocessing for the GENIA sentence splitter on PubMed abstracts, with minor tweaks for full-text documents.

sentences should be a string, with line breaks on sentence boundaries. Returns a similar string, but more correct.

Based on https://github.com/ninjin/geniass/blob/master/geniass-postproc.pl Which is (c) 2010 Sampo Pyysalo. No rights reserved, i.e. do whatever you like with this. Which draws in part on heuristics included in Yoshimasa Tsuruoka's medss.pl script.

WordTokenizers.punctuation_space_tokenize — Function.

punctuation_space_tokenize

Tokenizes by removing punctuation, unless it occurs inside of a word.

WordTokenizers.replace_til_no_change

WordTokenizers.rev_detokenize

WordTokenizers.rev_tokenize

WordTokenizers.rulebased_split_sentences

WordTokenizers.sentence_splitters

WordTokenizers.set_sentence_splitter — Function.

set_sentence_splitter(fun)

Call this to set the default sentence splitter to invoke the passed in function fun It will be used by split_sentences. Calling this will trigger recompilation of any functions that use split_sentences.

Calling set_sentence_splitter will give method overwritten warnings. They are expected, be worried if they do not occur

WordTokenizers.set_tokenizer — Function.

set_tokenizer(fun)

Call this to set the default tokenizer to invoke the passed in function fun It will be used by tokenize. Calling this will trigger recompilation of any functions that use tokenize.

Calling set_tokenizer will give method overwritten warnings. They are expected, be worried if they do not occur

WordTokenizers.spaces — Function.

spaces(::TokenBuffer)

If there is whitespace in the input, skip it, and flush the current token.

WordTokenizers.split_sentences

WordTokenizers.splits — Function.

splits(::TokenBuffer, [("cannot", 3), ("gimme", 3), ...])

Matches tokens that should be split at the given index. For example, cannot would be split into can and not.

WordTokenizers.suffixes — Function.

suffixes(::TokenBuffer, ["'ll", "'re", ...])

Matches tokens with suffixes, such as you're, that should be treated as separate tokens.

WordTokenizers.tokenize

WordTokenizers.tokenizers