Docstrings

JDBC.COLUMN_NO_NULLS

JDBC.COLUMN_NULLABLE

JDBC.COLUMN_NULLABLE_UNKNOWN

JDBC.Connection — Type.

Open a JDBC Connection to the specified host. The username and password can be optionally passed as a Dictionary props of the form Dict("user" => "username", "passwd" => "password"). The JDBC connector location can be optionally passed as connectorpath, if it is not added to the java class path.

Returns a JDBCConnection instance.

JDBC.Cursor

JDBC.DriverManager

JDBC.JCallableStatement

JDBC.JConnection

JDBC.JDBC

JDBC.JDBCConnection — Type.

Open a JDBC Connection to the specified host. The username and password can be optionally passed as a Dictionary props of the form Dict("user" => "username", "passwd" => "password"). The JDBC connector location can be optionally passed as connectorpath, if it is not added to the java class path.

Returns a JDBCConnection instance.

JDBC.JDBCCursor

JDBC.JDBCError

JDBC.JDBCRowIterator — Type.

Iterator to get rows of tables as array of tuples.

JDBC.JDBC_COLTYPE_ARRAY

JDBC.JDBC_COLTYPE_BIGINT

JDBC.JDBC_COLTYPE_BINARY

JDBC.JDBC_COLTYPE_BIT

JDBC.JDBC_COLTYPE_BLOB

JDBC.JDBC_COLTYPE_BOOLEAN

JDBC.JDBC_COLTYPE_CHAR

JDBC.JDBC_COLTYPE_CLOB

JDBC.JDBC_COLTYPE_DATALINK

JDBC.JDBC_COLTYPE_DATE

JDBC.JDBC_COLTYPE_DECIMAL

JDBC.JDBC_COLTYPE_DISTINCT

JDBC.JDBC_COLTYPE_DOUBLE

JDBC.JDBC_COLTYPE_FLOAT

JDBC.JDBC_COLTYPE_INTEGER

JDBC.JDBC_COLTYPE_JAVA_OBJECT

JDBC.JDBC_COLTYPE_LONGNVARCHAR

JDBC.JDBC_COLTYPE_LONGVARBINARY

JDBC.JDBC_COLTYPE_LONGVARCHAR

JDBC.JDBC_COLTYPE_NCHAR

JDBC.JDBC_COLTYPE_NCLOB

JDBC.JDBC_COLTYPE_NULL

JDBC.JDBC_COLTYPE_NUMERIC

JDBC.JDBC_COLTYPE_NVARCHAR

JDBC.JDBC_COLTYPE_OTHER

JDBC.JDBC_COLTYPE_REAL

JDBC.JDBC_COLTYPE_REF

JDBC.JDBC_COLTYPE_ROWID

JDBC.JDBC_COLTYPE_SMALLINT

JDBC.JDBC_COLTYPE_SQLXML

JDBC.JDBC_COLTYPE_STRUCT

JDBC.JDBC_COLTYPE_TIME

JDBC.JDBC_COLTYPE_TIMESTAMP

JDBC.JDBC_COLTYPE_TINYINT

JDBC.JDBC_COLTYPE_VARBINARY

JDBC.JDBC_COLTYPE_VARCHAR

JDBC.JPreparedStatement

JDBC.JResultSet

JDBC.JResultSetMetaData

JDBC.JStatement

JDBC.Source

JDBC.absolute! — Function.

absolute!(rs::JResultSet, row::Int)

Move the cursor to the specified row.

Args

rs: The JResultSet object.

row: The row to move to.

Returns

A boolean indicating success.

JDBC.beforeFirst! — Function.

beforeFirst!(rs::JResultSet)

Move the cursor to the front of the JResultSet, before the first row.

Args

rs: The JResultSet object.

Returns

nothing

JDBC.clearParameters — Function.

clearParameters(stmt::Union{JPreparedStatement, JCallableStatement})

Clears the currently held parameters in a JPreparedStatement object or a JCallableStatement object

Args

stmt: The JPreparedStatement object or JCallableStatement object

Returns

None

JDBC.colname

JDBC.colnames

JDBC.coltype

JDBC.coltypes

JDBC.column_types

JDBC.commit — Function.

commit(connection::JConnection)

Commits the transaction

Args

connection: The connection object

Returns

None

Commit any pending transaction to the database. Throws a JDBCError if connection is null.

Returns nothing.

JDBC.connect

JDBC.connection — Function.

Return the corresponding connection for a given cursor.

JDBC.createStatement — Function.

createStatement(connection::JConnection)

Initializes a Statement

Args

connection: The connection object

Returns

The JStatement object

JDBC.createstatement

JDBC.cursor — Function.

Create a new database cursor.

Returns a JDBCCursor instance.

JDBC.destroy

JDBC.eval

JDBC.execute — Function.

execute(stmt::(Union{JPreparedStatement, JCallableStatement}))

Executes the auery based on the Prepared Statement or Callable Statement

Args

stmt: The Prepared Statement or the Callable Statement object

Returns

A boolean indicating whether the execution was successful or not

execute(stmt::JStatement, query::AbstractString)

Executes the auery based on JStatement or any of its sub-types

Args

stmt: The JStatement object or any of its sub-types

query: The query to be executed

Returns

A boolean indicating whether the execution was successful or not

JDBC.execute! — Function.

Run a query on a database.

The results of the query are not returned by this function but are accessible through the cursor.

parameters can be any iterable of positional parameters, or of some T<:Associative for keyword/named parameters.

Throws a JDBCError if query caused an error, cursor is not initialized or connection is null.

Returns nothing.

JDBC.executeQuery — Function.

executeQuery(stmt::JStatement, query::AbstractString)

Executes the auery and returns the results as a JResultSet object.

Args

stmt: The Statement object

query: The query to be executed

Returns

The result set as a JResultSet object

executeQuery(stmt::Union{JPreparedStatement, JCallableStatement})

Executes the auery based on a JPreparedStatement object or a JCallableStatement object

Args

stmt: The JPreparedStatement object or JCallableStatement object

Returns

The result set as a JResultSet object

JDBC.executeUpdate — Function.

executeUpdate(stmt::JStatement, query::AbstractString)

Executes the update auery and returns the status of the execution of the query

Args

stmt: The Statement object

query: The query to be executed

Returns

An integer representing the status of the execution

executeUpdate(stmt::Union{JPreparedStatement, JCallableStatement})

Executes the update auery based on a JPreparedStatement object or a JCallableStatement object

Args

stmt: The JPreparedStatement object or JCallableStatement object

Returns

An integer indicating the status of the execution of the query

JDBC.getBigDecimal

JDBC.getBoolean

JDBC.getByte

JDBC.getColumnCount — Function.

getColumnCount(rsmd::JResultSetMetaData)

Returns the number of columns based on the JResultSetMetaData object

Args

rsmd: The JResultSetMetaData object

Returns

The number of columns.

JDBC.getColumnName — Function.

getColumnName(rsmd::JResultSetMetaData, col::Integer)

Returns the column's name based on the JResultSetMetaData object and the column number

Args

rsmd: The JResultSetMetaData object

col: The column number

Returns

The column name

JDBC.getColumnType — Function.

getColumnType(rsmd::JResultSetMetaData, col::Integer)

Returns the column's data type based on the JResultSetMetaData object and the column number

Args

rsmd: The JResultSetMetaData object

col: The column number

Returns

The column type as an integer

JDBC.getDate — Function.

getDate(rs::Union{JResultSet, JCallableStatement}, fld::AbstractString)

Returns the Date object based on the result set or a callable statement. The value is extracted based on the column name.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column name

Returns

The Date object.

getDate(rs::Union{JResultSet, JCallableStatement}, fld::Integer)

Returns the Date object based on the result set or a callable statement. The value is extracted based on the column number.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column number

Returns

The Date object.

JDBC.getDouble

JDBC.getFloat

JDBC.getInt

JDBC.getLong

JDBC.getMetaData — Function.

getMetaData(rs::JResultSet)

Returns information about the types and properties of the columns in the ResultSet object

Args

stmt: The JResultSet object

Returns

The JResultSetMetaData object.

JDBC.getNString

JDBC.getResultSet — Function.

getResultSet(stmt::JStatement)

Returns the result set based on the previous execution of the query based on a JStatement

Args

stmt: The JStatement object

Returns

The JResultSet object.

JDBC.getShort

JDBC.getString

JDBC.getTableMetaData — Function.

Get the metadata (column name and type) for each column of the table in the result set rs.

Returns an array of (column name, column type) tuples.

JDBC.getTime — Function.

getTime(rs::Union{JResultSet, JCallableStatement}, fld::AbstractString)

Returns the Time object based on the result set or a callable statement. The value is extracted based on the column name.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column name

Returns

The Time object.

getTime(rs::Union{JResultSet, JCallableStatement}, fld::Integer)

Returns the Time object based on the result set or a callable statement. The value is extracted based on the column number.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column number

Returns

The Time object.

JDBC.getTimestamp — Function.

getTimestamp(rs::Union{JResultSet, JCallableStatement}, fld::AbstractString)

Returns the Timestamp object based on the result set or a callable statement. The value is extracted based on the column name.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column name

Returns

The Timestamp object.

getTimestamp(rs::@Union{JResultSet, JCallableStatement}, fld::Integer)

Returns the Timestamp object based on the result set or a callable statement. The value is extracted based on the column number.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column number

Returns

The Timestamp object.

JDBC.getURL

JDBC.get_method_dict

JDBC.include

JDBC.init

JDBC.isNullable

JDBC.isdone

JDBC.isopen — Function.

Returns a boolean indicating whether connection conn is open.

JDBC.jdbc_get_method

JDBC.load

JDBC.ncols

JDBC.prepareCall — Function.

prepareCall(connection::JConnection, query::AbstractString)

Prepares the Callable Statement for the given query

Args

connection: The connection object

query: The query string

Returns

The JCallableStatement object

JDBC.prepareStatement — Function.

prepareStatement(connection::JConnection, query::AbstractString)

Prepares the Statement for the given query

Args

connection: The connection object

query: The query string

Returns

The JPreparedStatement object

JDBC.pullfield

JDBC.rollback — Function.

rollback(connection::JConnection)

Rolls back the transactions.

Args

connection: The connection object

Returns

None

Roll back to the start of any pending transaction. Throws a JDBCError if connection is null.

Returns nothing.

JDBC.rows — Function.

Create a row iterator.

This method returns an instance of an iterator type which returns one row on each iteration. Each row returnes a Tuple{...}.

Throws a JDBCError if execute! was not called on the cursor or connection is null.

Returns a JDBCRowIterator instance.

JDBC.setAutoCommit — Function.

setAutoCommit(connection::JConnection, x::Bool)

Set the Auto Commit flag to either true or false. If set to false, commit has to be called explicitly

Args

connection: The connection object

Returns

None

JDBC.setBigDecimal

JDBC.setBoolean

JDBC.setByte

JDBC.setDouble

JDBC.setFetchSize — Function.

setFetchSize(stmt::Union{JStatement, JPreparedStatement, JCallableStatement }, x::Integer)

Sets the fetch size in a JStatement or a JPreparedStatement object or a JCallableStatement object. The number of records that are returned in subsequent query executions are determined by what is set here.

Args

stmt: The JPreparedStatement object or JCallableStatement object

x: The number of records to be returned

Returns

None

JDBC.setFloat

JDBC.setInt

JDBC.setLong

JDBC.setNString

JDBC.setShort

JDBC.setString

JDBC.setURL

JDBC.usedriver

JDBC.wasNull