Docstrings

FilePathsBase.@LOCAL — Macro.

@LOCAL(filespec)

Construct an absolute path to filespec relative to the source file containing the macro call.

FilePathsBase.@__FILEPATH__ — Macro.

@__FILEPATH__ -> AbstractPath

@FILEPATH expands to a path with the absolute file path of the file containing the macro. Returns an empty Path if run from a REPL or if evaluated by julia -e <expr>.

FilePathsBase.@__PATH__ — Macro.

@__PATH__ -> AbstractPath

@PATH expands to a path with the directory part of the absolute path of the file containing the macro. Returns an empty Path if run from a REPL or if evaluated by julia -e <expr>.

FilePathsBase.@p_str — Macro.

@p_str -> Path

Constructs a Path (platform specific subtype of AbstractPath), such as p"~/.juliarc.jl".

FilePathsBase.AbstractPath

FilePathsBase.Cgroup

FilePathsBase.Cpasswd

FilePathsBase.DATA_SUFFIX

FilePathsBase.EXEC

FilePathsBase.FILEMODE_TABLE

FilePathsBase.FilePathsBase

FilePathsBase.GROUP_COEFF

FilePathsBase.Group

FilePathsBase.Mode — Type.

The following file contains simple abstractions to make working with posix permissions easier in julia.

For now we're focused on user permissions.

A lot of the low level permissions code below and the corresponding constants have been translated from cpython's Lib/stat.py file.

https://github.com/python/cpython/blob/master/Lib/stat.py

FilePathsBase.OTHER_COEFF

FilePathsBase.PATH_TYPES

FilePathsBase.POSIX_PATH_ABSOLUTE_RE

FilePathsBase.POSIX_PATH_DIRECTORY_RE

FilePathsBase.POSIX_PATH_DIR_SPLITTER

FilePathsBase.POSIX_PATH_EXT_SPLITTER

FilePathsBase.POSIX_PATH_SEPARATOR

FilePathsBase.POSIX_PATH_SEPARATOR_RE

FilePathsBase.Path — Function.

Path()
Path(path::AbstractPath)
Path(path::Tuple)
Path(path::AbstractString)

Responsible for creating the appropriate platform specific path (e.g., PosixPath and WindowsPath for Unix and Windows systems respectively)

FilePathsBase.PosixPath

FilePathsBase.READ

FilePathsBase.S_ENFMT

FilePathsBase.S_IEXEC

FilePathsBase.S_IFBLK

FilePathsBase.S_IFCHR

FilePathsBase.S_IFDIR

FilePathsBase.S_IFIFO

FilePathsBase.S_IFLNK

FilePathsBase.S_IFREG

FilePathsBase.S_IFSOCK

FilePathsBase.S_IREAD

FilePathsBase.S_IRGRP

FilePathsBase.S_IROTH

FilePathsBase.S_IRUSR

FilePathsBase.S_IRWXG

FilePathsBase.S_IRWXO

FilePathsBase.S_IRWXU

FilePathsBase.S_ISGID

FilePathsBase.S_ISUID

FilePathsBase.S_ISVTX

FilePathsBase.S_IWGRP

FilePathsBase.S_IWOTH

FilePathsBase.S_IWRITE

FilePathsBase.S_IWUSR

FilePathsBase.S_IXGRP

FilePathsBase.S_IXOTH

FilePathsBase.S_IXUSR

FilePathsBase.Status

FilePathsBase.USER_COEFF

FilePathsBase.User

FilePathsBase.WIN_PATH_ABSOLUTE_RE

FilePathsBase.WIN_PATH_DIRECTORY_RE

FilePathsBase.WIN_PATH_DIR_SPLITTER

FilePathsBase.WIN_PATH_EXT_SPLITTER

FilePathsBase.WIN_PATH_SEPARATOR

FilePathsBase.WIN_PATH_SEPARATOR_RE

FilePathsBase.WRITE

FilePathsBase.WindowsPath

FilePathsBase.__init__

FilePathsBase._datasize

FilePathsBase._meta — Function.

Return the portion of the file's mode that describes the file type.

FilePathsBase._mode — Function.

Return the portion of the file's mode that can be set by os.chmod().

FilePathsBase._win_splitdrive

FilePathsBase.abspath

FilePathsBase.anchor

Base.Filesystem.chown — Function.

chown(path::AbstractPath, user::AbstractString, group::AbstractString; recursive=false)

Change the user and group of the path.

FilePathsBase.created — Function.

created(path::AbstractPath) -> DateTime

Returns the creation date for the path.

Example

julia> created(p"src/FilePathsBase.jl")
2017-06-20T04:01:09

FilePathsBase.cwd

FilePathsBase.dirname

FilePathsBase.drive

FilePathsBase.eval

FilePathsBase.executable

FilePathsBase.exists — Function.

exists(path::AbstractPath) -> Bool

Returns whether the path actually exists on the system.

FilePathsBase.expanduser

FilePathsBase.extension — Function.

extension(path::AbstractPath) -> AbstractString

Extracts the last extension from a filename if there any, otherwise it returns an empty string.

Example

julia> extension(p"~/repos/FilePathsBase.jl/src/FilePathsBase.jl")
"jl"

FilePathsBase.extensions — Function.

extensions(path::AbstractPath) -> AbstractString

Extracts all extensions from a filename if there any, otherwise it returns an empty string.

Example

julia> extensions(p"~/repos/FilePathsBase.jl/src/FilePathsBase.jl.bak")
2-element Array{SubString{String},1}:
 "jl"
 "bak"

FilePathsBase.filemode

FilePathsBase.filename — Function.

filename(path::AbstractPath) -> AbstractString

Extracts the basename without any extensions.

Example

julia> filename(p"~/repos/FilePathsBase.jl/src/FilePathsBase.jl")
"FilePathsBase"

FilePathsBase.hasparent — Function.

hasparent(path::AbstractPath) -> Bool

Returns whether there is a parent directory component to the supplied path.

FilePathsBase.home

FilePathsBase.include

FilePathsBase.isabs

FilePathsBase.isabspath

FilePathsBase.isexecutable — Function.

isexecutable(path::AbstractPath) -> Bool

Returns whether the path is executable for the current user.

FilePathsBase.ispath

FilePathsBase.ispathtype

FilePathsBase.mkpath

FilePathsBase.mktmp

FilePathsBase.mktmpdir

FilePathsBase.mode — Function.

mode(path::AbstractPath) -> Mode

Returns the Mode for the specified path.

Example

julia> mode(p"src/FilePathsBase.jl")
-rw-r--r--

FilePathsBase.modified — Function.

modified(path::AbstractPath) -> DateTime

Returns the last modified date for the path.

Example

julia> modified(p"src/FilePathsBase.jl")
2017-06-20T04:01:09

FilePathsBase.move

FilePathsBase.mv

FilePathsBase.normpath

FilePathsBase.parents — Function.

parents{T<:AbstractPath}(path::T) -> Array{T}

Example

julia> parents(p"~/.julia/v0.6/REQUIRE") 3-element Array{FilePathsBase.PosixPath,1}: p"~" p"~/.julia" p"~/.julia/v0.6"

Throws

ErrorException: if path doesn't have a parent

FilePathsBase.parts

FilePathsBase.raw

FilePathsBase.readable

FilePathsBase.realpath

FilePathsBase.register

FilePathsBase.relative — Function.

relative{T<:AbstractPath}(path::T, start::T=T("."))

Creates a relative path from either the current directory or an arbitrary start directory.

FilePathsBase.relpath

FilePathsBase.remove

FilePathsBase.rm

FilePathsBase.root

FilePathsBase.tmpdir

FilePathsBase.tmpname

FilePathsBase.writable