Docstrings

Meshing.AbstractMeshingAlgorithm

Meshing.MarchingCubes

Meshing.MarchingTetrahedra

Meshing.Meshing

Meshing.VoxelIndices — Type.

Voxel corner and edge indexing conventions

    Z
    |

    5------5------6              Extra edges not drawn
   /|            /|              -----------
  8 |           6 |              - face diagonals
 /  9          /  10                - 13: 1 to 3
8------7------7   |                 - 14: 1 to 8
|   |         |   |                 - 15: 1 to 6
|   1------1--|---2  -- Y           - 16: 5 to 7
12 /          11 /                  - 17: 2 to 7
| 4           | 2                   - 18: 4 to 7
|/            |/                 - body diagonal
4------3------3                     - 19: 1 to 7

/ X

Meshing._correct_vertices! — Function.

The marchingTetrahedra function returns vertices on the (1-based) indices of the SDF's data, ignoring its actual bounds. This function adjusts the vertices in place so that they correspond to points within the SDF bounds.

Meshing.edge_table

Meshing.eval

Meshing.getVertId — Function.

Gets the vertex ID, adding it to the vertex dictionary if not already present.

Meshing.hasFaces — Function.

Checks if a voxel has faces. Should be false for most voxels. This function should be made as fast as possible.

Meshing.include

Meshing.isosurface

Meshing.marchingTetrahedra — Function.

Given a 3D array and an isovalue, extracts a mesh represention of the an approximate isosurface by the method of marching tetrahedra.

Meshing.marching_cubes — Function.

marching_cubes(sdf::SignedDistanceField, [iso = 0.0,] [MT = HomogenousMesh{Point{3,Float64},Face{3,Int}}])

Construct a HomogenousMesh from a SignedDistanceField using the marching cubes algorithm. This method is faster than Marching Tetrahedra and generates fewer vertices and faces (about 1/4 as many). However it may generate non-manifold meshes, while Marching Tetrahedra guarentees a manifold mesh.

Meshing.procVox — Function.

Processes a voxel, adding any new vertices and faces to the given containers as necessary.

Meshing.tetIx — Function.

Determines which case in the triangle table we are dealing with

Meshing.tri_table

Meshing.vertId — Function.

Determines a unique integer ID associated with the edge. This is used as a key in the vertex dictionary. It needs to be both unambiguous (no two edges get the same index) and unique (every edge gets the same ID regardless of which of its neighboring voxels is asking for it) in order for vertex sharing to be implemented properly.

Meshing.vertPos — Function.

Assuming an edge crossing, determines the point in space at which it occurs. eps represents the "bump" factor to keep vertices away from voxel corners (thereby preventing degeneracies).

Meshing.vertex_interp

Meshing.voxEdgeId — Function.

Given a sub-tetrahedron case and a tetrahedron edge ID, determines the corresponding voxel edge ID.