Docstrings

GraphRecipes.GraphPlot

GraphRecipes.GraphRecipes

GraphRecipes._graph_funcs

GraphRecipes._graph_inputs

GraphRecipes.add_ast

GraphRecipes.add_subs!

GraphRecipes.adjlist_and_degrees

GraphRecipes.arc_diagram

GraphRecipes.arcdiagram_limits

GraphRecipes.arcshape

GraphRecipes.buchheim_graph

GraphRecipes.by_axis_local_stress_graph

GraphRecipes.chord_diagram

GraphRecipes.compute_laplacian

GraphRecipes.compute_tree_layers

GraphRecipes.compute_tree_layers2

GraphRecipes.directed_curve — Function.

This function builds a BezierCurve which leaves point p vertically upwards and arrives point q vertically upwards. It may create a loop if necessary. It assumes the view is [0,1]. That can be modified using the xview and yview keyword arguments (default: 0:1).

GraphRecipes.estimate_distance

GraphRecipes.eval

GraphRecipes.get_adjacency_list

GraphRecipes.get_adjacency_matrix

GraphRecipes.get_source_destiny_weight

GraphRecipes.graphplot

GraphRecipes.graphplot!

GraphRecipes.include

GraphRecipes.isnothing

GraphRecipes.make_symmetric

GraphRecipes.norm_ij

GraphRecipes.partialcircle — Function.

get an array of tuples of points on a circle with radius r

GraphRecipes.prepare_graph_inputs

GraphRecipes.random_control_point — Function.

Randomly pick a point to be the center control point of a bezier curve, which is both equidistant between the endpoints and normally distributed around the midpoint.

GraphRecipes.shift_children!

GraphRecipes.shorten_segment

GraphRecipes.spectral_graph

GraphRecipes.stress

GraphRecipes.tree_graph