Docstrings

MLDataUtils.BalancedObs

MLDataUtils.BatchView

MLDataUtils.BufferGetObs

MLDataUtils.DataSubset

MLDataUtils.FeatureNormalizer

MLDataUtils.FoldsView

MLDataUtils.LabelEnc

MLDataUtils.MLDataUtils

MLDataUtils.ObsDim

MLDataUtils.ObsView

MLDataUtils.RandomBatches

MLDataUtils.RandomObs

MLDataUtils.SlidingWindow

MLDataUtils._throw_table_error

MLDataUtils.batchsize

MLDataUtils.batchview

MLDataUtils.center! — Function.

μ = center!(X[, μ, obsdim])

or

μ = center!(D[, colnames, μ])

where X is of type Matrix or Vector and D of type DataFrame.

Center X along obsdim around the corresponding entry in the vector μ. If μ is not specified then it defaults to the feature specific means.

For DataFrames, obsdim is obsolete and centering is done column wise. Instead the vector colnames allows to specify which columns to center. If colnames is not provided all columns of type T<:Real are centered.

Example:

X = rand(4, 100)
D = DataFrame(A=rand(10), B=collect(1:10), C=[string(x) for x in 1:10])

μ = center!(X, obsdim=2)
μ = center!(X, ObsDim.First())
μ = center!(D)
μ = center!(D, [:A, :B])

MLDataUtils.classify

MLDataUtils.classify!

MLDataUtils.convertlabel

MLDataUtils.convertlabelview

MLDataUtils.datasubset

MLDataUtils.downsample

MLDataUtils.eachbatch

MLDataUtils.eachobs

MLDataUtils.eachtarget

MLDataUtils.eval

MLDataUtils.expand_poly — Function.

X = DataUtils.expand_poly(x; degree = 5)

Performs a polynomial basis expansion of the given degree for the vector x. The return value X is a matrix of size (degree, length(x)).

Note: all the features of X are centered and rescaled.

MLDataUtils.fit

MLDataUtils.getobs

MLDataUtils.getobs!

MLDataUtils.include

MLDataUtils.ind2label

MLDataUtils.islabelenc

MLDataUtils.isneglabel

MLDataUtils.isposlabel

MLDataUtils.kfolds

MLDataUtils.label

MLDataUtils.label2ind

MLDataUtils.labelenc

MLDataUtils.labelfreq

MLDataUtils.labelfreq!

MLDataUtils.labelmap

MLDataUtils.labelmap!

MLDataUtils.labelmap2vec

MLDataUtils.labeltype

MLDataUtils.leaveout

MLDataUtils.load_iris — Function.

X, y, vars = load_iris(n)

Loads the first n (of 150) observations from the Iris flower data set introduced by Ronald Fisher (1936). The 4 by n matrix X contains the numeric measurements, in which each individual column denotes an observation. The vector y contains the class labels as strings. The optional vector vars contains the names of the features (i.e. rows of X)

[1] Fisher, Ronald A. "The use of multiple measurements in taxonomic problems." Annals of eugenics 7.2 (1936): 179-188.

MLDataUtils.load_line — Function.

x, y, vars = load_line()

Loads an artificial example dataset for a noisy line. It is particularly useful to explain under- and overfitting. The vector x contains 11 equally spaced points between 0 and 1 The vector y contains x ./ 2 + 1 plus some gaussian noise The optional vector vars contains descriptive names for x and y

MLDataUtils.load_poly — Function.

x, y, vars = load_poly()

Loads an artificial example dataset for a noisy quadratic function. It is particularly useful to explain under- and overfitting. The vector x contains 50 points between 0 and 4 The vector y contains 2.6 * x^2 + .8 * x plus some gaussian noise The optional vector vars contains descriptive names for x and y

MLDataUtils.load_sin — Function.

x, y, vars = load_sin()

Loads an artificial example dataset for a noisy sin. It is particularly useful to explain under- and overfitting. The vector x contains equally spaced points between 0 and 2π The vector y contains sin(x) plus some gaussian noise The optional vector vars contains descriptive names for x and y

MLDataUtils.load_spiral — Function.

x, y, vars = load_spiral()`

Loads an artificial example dataset for a noisy spiral function. It is particularly useful to explain representation learning and nonlinearity. The matrix x contains 194 points between 0 and 6.5 lying on the spiral. The vector y contains the corresponding labels, i.e "ones" or "zeros". The optional vector vars contains descriptive names for x and y

MLDataUtils.neglabel

MLDataUtils.nlabel

MLDataUtils.nobs

MLDataUtils.noisy_function — Function.

x, y = noisy_function(fun, x; noise = 0.01, f_rand = randn)

Generates a noisy response y for the given function fun by adding noise .* f_randn(length(x)) to the result of fun(x).

MLDataUtils.noisy_poly — Function.

x, y = noisy_poly(coef, x; noise = 0.01, f_rand = randn)

Generates a noisy response for a polynomial of degree length(coef) using the vector x as input and adding noise .* f_randn(length(x)) to the result. The vector coef contains the coefficients for the terms of the polynome. The first element of coef denotes the coefficient for the term with the highest degree, while the last element of coef denotes the intercept.

MLDataUtils.noisy_sin — Function.

x, y = noisy_sin(n, start, stop; noise = 0.3, f_rand = randn)

Generates n noisy equally spaces samples of a sinus from start to stop by adding noise .* f_randn(length(x)) to the result of fun(x).

MLDataUtils.noisy_spiral — Function.

x, y = noisy_spiral(n, a, theta, b; noise = 0.01, f_rand = randn)

Generates n noisy responses for a spiral with two labels. Uses the radius, angle and scaling arguments to space the points in 2D space and adding noise .* f_randn(n) to the response.

MLDataUtils.obsview

MLDataUtils.oversample

MLDataUtils.poslabel

MLDataUtils.predict

MLDataUtils.predict!

MLDataUtils.randobs

MLDataUtils.rescale! — Function.

μ, σ = rescale!(X[, μ, σ, obsdim])

or

μ, σ = rescale!(D[, colnames, μ, σ])

where X is of type Matrix or Vector and D of type DataFrame.

Center X along obsdim around the corresponding entry in the vector μ and then rescale each feature using the corresponding entry in the vector σ.

For DataFrames, obsdim is obsolete and centering is done column wise. The vector colnames allows to specify which columns to center. If colnames is not provided all columns of type T<:Real are centered.

Example:

X = rand(4, 100)
D = DataFrame(A=rand(10), B=collect(1:10), C=[string(x) for x in 1:10])

μ, σ = rescale!(X, obsdim=2)
μ, σ = rescale!(X, ObsDim.First())
μ, σ = rescale!(D)
μ, σ = rescale!(D, [:A, :B])

MLDataUtils.shuffleobs

MLDataUtils.slidingwindow

MLDataUtils.splitobs

MLDataUtils.stratifiedobs

MLDataUtils.targets

MLDataUtils.undersample

MLDataUtils.upsample