Docstrings

GeoStatsDevTools.+

GeoStatsDevTools.-

GeoStatsDevTools.AbstractMixtureModel

GeoStatsDevTools.AbstractMvNormal

GeoStatsDevTools.AbstractNeighborSearcher — Type.

AbstractNeighborSearcher

A method for searching neighbors in a spatial object given a reference point.

GeoStatsDevTools.AbstractNeighborhood — Type.

AbstractNeighborhood{D}

A neighborhood on a spatial domain of type D.

GeoStatsDevTools.AbstractPartitioner — Type.

AbstractPartitioner

A method for partitioning spatial data.

GeoStatsDevTools.AbstractPath — Type.

AbstractPath

A path on a spatial domain of type D.

GeoStatsDevTools.Arcsine

GeoStatsDevTools.BallNeighborhood — Type.

BallNeighborhood(domain, radius)

A ball neighborhood of a given radius on a spatial domain.

GeoStatsDevTools.Bernoulli

GeoStatsDevTools.Beta

GeoStatsDevTools.BetaBinomial

GeoStatsDevTools.BetaPrime

GeoStatsDevTools.Binomial

GeoStatsDevTools.Biweight

GeoStatsDevTools.Categorical

GeoStatsDevTools.Cauchy

GeoStatsDevTools.Chi

GeoStatsDevTools.Chisq

GeoStatsDevTools.Continuous

GeoStatsDevTools.ContinuousDistribution

GeoStatsDevTools.ContinuousMatrixDistribution

GeoStatsDevTools.ContinuousMultivariateDistribution

GeoStatsDevTools.ContinuousUnivariateDistribution

GeoStatsDevTools.Cosine

GeoStatsDevTools.CylinderNeighborhood — Type.

CylinderNeighborhood(domain, radius, height)

A cylinder neighborhood with a given radius and height on a spatial domain.

Notes

The height parameter is only half of the actual cylinder height.

GeoStatsDevTools.DiagNormal

GeoStatsDevTools.DiagNormalCanon

GeoStatsDevTools.DirectionPartitioner — Type.

DirectionPartitioner(direction; atol=20., btol=.95)

A method for partitioning spatial data along a given direction with angle tolerance atol in degrees and bandwidth tolerance btol.

      ________________
     /        | btol
    /         |             ILLUSTRATION OF DIRECTION TOLERANCES
    ----------------->
    \ ) atol
     \________________

GeoStatsDevTools.Dirichlet

GeoStatsDevTools.DirichletMultinomial

GeoStatsDevTools.Discrete

GeoStatsDevTools.DiscreteDistribution

GeoStatsDevTools.DiscreteMatrixDistribution

GeoStatsDevTools.DiscreteMultivariateDistribution

GeoStatsDevTools.DiscreteUniform

GeoStatsDevTools.DiscreteUnivariateDistribution

GeoStatsDevTools.Distribution

GeoStatsDevTools.Distributions

GeoStatsDevTools.DoubleExponential

GeoStatsDevTools.EdgeworthMean

GeoStatsDevTools.EdgeworthSum

GeoStatsDevTools.EdgeworthZ

GeoStatsDevTools.Ellipsoidal — Type.

Ellipsoidal(semiaxes, angles)

A distance defined by an ellipsoid with given semiaxes and rotation angles.

For 2D ellipsoids, there are two semiaxes and one rotation angle.

For 3D ellipsoids, there are three semiaxes and three rotation angles.

Examples

2D ellipsoid making 45ᵒ with the horizontal axis:

julia> Ellipsoidal([1.0,0.5], [π/2])

3D ellipsoid rotated by 45ᵒ in the xy plane:

julia> Ellipsoidal([1.0,0.5,0.5], [π/2,0.0,0.0])

GeoStatsDevTools.EmpiricalDistribution — Type.

EmpiricalDistribution(values)

An empirical distribution holding continuous values.

GeoStatsDevTools.EmpiricalUnivariateDistribution

GeoStatsDevTools.Epanechnikov

GeoStatsDevTools.Erlang

GeoStatsDevTools.Estimator

GeoStatsDevTools.Exponential

GeoStatsDevTools.FDist

GeoStatsDevTools.FisherNoncentralHypergeometric

GeoStatsDevTools.Frechet

GeoStatsDevTools.FullNormal

GeoStatsDevTools.FullNormalCanon

GeoStatsDevTools.Gamma

GeoStatsDevTools.GeneralizedExtremeValue

GeoStatsDevTools.GeneralizedPareto

GeoStatsDevTools.GeoDataFrame — Type.

GeoDataFrame(data, coordnames)

A dataframe object data with additional metadata for tracking the columns coordnames that represent spatial coordinates.

Examples

If the data was already loaded in a normal DataFrame data, and there exists columns named x, y and z, wrap the data and specify the column names:

julia> GeoDataFrame(data, [:x,:y,:z])

Alternatively, load the data directly into a GeoDataFrame object by using the method readgeotable.

Notes

This type is a lightweight wrapper over Julia's DataFrame types. No additional storage is required other than a vector of symbols with the columns names representing spatial coordinates.

GeoStatsDevTools.GeoStatsDevTools

GeoStatsDevTools.Geometric

GeoStatsDevTools.Gumbel

GeoStatsDevTools.HierarchicalPartitioner — Type.

HierarchicalPartitioner(first, second)

A partitioning method in which a first partition is applied and then a second partition is applied to each subset of the first.

GeoStatsDevTools.Hypergeometric

GeoStatsDevTools.InverseGamma

GeoStatsDevTools.InverseGaussian

GeoStatsDevTools.InverseWishart

GeoStatsDevTools.IsoNormal

GeoStatsDevTools.IsoNormalCanon

GeoStatsDevTools.KSDist

GeoStatsDevTools.KSOneSided

GeoStatsDevTools.Kolmogorov

GeoStatsDevTools.Laplace

GeoStatsDevTools.Levy

GeoStatsDevTools.LocalNeighborSearcher — Type.

LocalNeighborSearcher(domain, K, neighborhood, path)

A search method that finds at most K neighbors in neighborhood of domain with a search path.

GeoStatsDevTools.LocationScale

GeoStatsDevTools.LogNormal

GeoStatsDevTools.Logistic

GeoStatsDevTools.MLEstimator

GeoStatsDevTools.MatrixDistribution

GeoStatsDevTools.Matrixvariate

GeoStatsDevTools.MixtureModel

GeoStatsDevTools.Multinomial

GeoStatsDevTools.Multivariate

GeoStatsDevTools.MultivariateDistribution

GeoStatsDevTools.MultivariateMixture

GeoStatsDevTools.MultivariateNormal

GeoStatsDevTools.MvLogNormal

GeoStatsDevTools.MvNormal

GeoStatsDevTools.MvNormalCanon

GeoStatsDevTools.MvNormalKnownCov

GeoStatsDevTools.MvTDist

GeoStatsDevTools.NearestNeighborSearcher — Type.

NearestNeighborSearcher(domain, K, [metric])

A search method that finds at most K nearest neighbors in domain according to metric (default to Euclidean()).

GeoStatsDevTools.NegativeBinomial

GeoStatsDevTools.NonMatrixDistribution

GeoStatsDevTools.NoncentralBeta

GeoStatsDevTools.NoncentralChisq

GeoStatsDevTools.NoncentralF

GeoStatsDevTools.NoncentralHypergeometric

GeoStatsDevTools.NoncentralT

GeoStatsDevTools.Normal

GeoStatsDevTools.NormalCanon

GeoStatsDevTools.NormalInverseGaussian

GeoStatsDevTools.Pareto

GeoStatsDevTools.PlanePartitioner — Type.

PlanePartitioner(normal; tol=1e-6)

A method for partitioning spatial data into a family of hyperplanes defined by a normal direction. Two points x and y belong to the same hyperplane when (x - y) ⋅ normal < tol.

GeoStatsDevTools.PointSet — Type.

PointSet(coords)

A set of points with coordinate matrix coords. The number of rows of the matrix is the dimensionality of the domain whereas the number of columns is the number of points in the set.

GeoStatsDevTools.PointSetData — Type.

PointSetData(data, coords)

Spatial data georeferenced with coordinate matrix coords. The data argument is a dictionary mapping variable names to Julia arrays with the actual data.

See also: PointSet

GeoStatsDevTools.Poisson

GeoStatsDevTools.PoissonBinomial

GeoStatsDevTools.QQPair

GeoStatsDevTools.RandomPath — Type.

RandomPath(domain)

A random path on a spatial domain.

GeoStatsDevTools.Rayleigh

GeoStatsDevTools.RealInterval

GeoStatsDevTools.RegularGrid — Type.

RegularGrid(dims, origin, spacing)

A regular grid with dimensions dims, lower left corner at origin and cell spacing spacing. The three arguments must have the same length.

RegularGrid(start, finish, dims=dims)

Alternatively, construct a regular grid from a start point (lower left) to a finish point (upper right), optionally passing dims.

RegularGrid{T}(dims)
RegularGrid{T}(dim1, dim2, ...)

Finally, a regular grid can be constructed by only passing the dimensions dims as a tuple, or by passing each dimension dim1, dim2, ... separately. In this case, the origin and spacing default to (0,0,...) and (1,1,...).

Examples

Create a 3D grid with 100x100x50 locations:

julia> RegularGrid{Float64}(100,100,50)

Create a 2D grid with 100x100 locations and origin at (10.,20.) units:

julia> RegularGrid((100,100),(10.,20.),(1.,1.))

Create a 1D grid from -1 to 1 with 100 locations:

julia> RegularGrid((-1.,),(1.,), dims=(100,))

GeoStatsDevTools.RegularGridData — Type.

RegularGridData(data, origin, spacing)

Regularly spaced data georeferenced with origin and spacing. The data argument is a dictionary mapping variable names to Julia arrays with the actual data.

NaN or missing values in the Julia arrays are interpreted as non-valid. They can be used to mask the variables on the grid.

Examples

Given poro and perm two 2-dimensional Julia arrays containing values of porosity and permeability, the following code can be used to georeference the data:

julia> data = Dict(:porosity => poro, :permeability => perm)
julia> RegularGridData(data, [0.,0.,0.], [1.,1.,1.])

Alternatively, one can omit origin and spacing for default values of zeros and ones:

julia> RegularGridData{Float64}(data)

See also: RegularGrid

GeoStatsDevTools.Sampleable

GeoStatsDevTools.Semicircle

GeoStatsDevTools.ShiftedPath — Type.

ShiftedPath(path, offset)

Produces the same results of path but shifted by an offset that can be positive or negative.

GeoStatsDevTools.SimplePath — Type.

SimplePath(domain)

A simple (or default) path on a spatial domain.

GeoStatsDevTools.Skellam

GeoStatsDevTools.SourcePath — Type.

SourcePath(domain, sources)

A path over a spatial domain that starts at given source locations sources and progresses outwards.

GeoStatsDevTools.SpatialPartition — Type.

SpatialPartition

A partition of spatial data.

GeoStatsDevTools.SpatialPredicatePartitioner — Type.

SpatialPredicatePartitioner(predicate)

A method for partitioning spatial data based on a predicate. Given two coordinates x and y, the value predicate(x,y) informs whether or not the points belong to the same subset.

GeoStatsDevTools.SpatialStatistic — Type.

SpatialStatistic{D}

A spatial statistic on a spatial domain of type D.

GeoStatsDevTools.StructuredGridData — Type.

StructuredGridData(data, X, Y, Z, ...)

Data spatially distributed on a structured grid where points are georeferenced by coordinates X, Y, Z, ...

The data argument is a dictionary mapping variable names to Julia arrays with the actual data.

Examples

A very popular structured grid data format is NetCDF. Given 2D arrays LAT and LON with coordinates and arrays with climate data precipitation, temperature, the following code can be used to construct a structured grid:

julia> data = Dict(:precipitation => precipitation, :temperature => temperature)
julia> StructuredGridData(data, LAT, LON)

GeoStatsDevTools.SufficientStats

GeoStatsDevTools.SymTriangularDist

GeoStatsDevTools.TDist

GeoStatsDevTools.TriangularDist

GeoStatsDevTools.Triweight

GeoStatsDevTools.Truncated

GeoStatsDevTools.TruncatedNormal

GeoStatsDevTools.Uniform

GeoStatsDevTools.UniformPartitioner — Type.

UniformPartitioner(k, [shuffle])

A method for partitioning spatial data uniformly into k subsets of approximately equal size. Optionally shuffle the data (default to true).

GeoStatsDevTools.Univariate

GeoStatsDevTools.UnivariateDistribution

GeoStatsDevTools.UnivariateGMM

GeoStatsDevTools.UnivariateMixture

GeoStatsDevTools.ValueSupport

GeoStatsDevTools.VariateForm

GeoStatsDevTools.VonMises

GeoStatsDevTools.VonMisesFisher

GeoStatsDevTools.WalleniusNoncentralHypergeometric

GeoStatsDevTools.Weibull

GeoStatsDevTools.Wishart

GeoStatsDevTools.ZeroMeanDiagNormal

GeoStatsDevTools.ZeroMeanDiagNormalCanon

GeoStatsDevTools.ZeroMeanFullNormal

GeoStatsDevTools.ZeroMeanFullNormalCanon

GeoStatsDevTools.ZeroMeanIsoNormal

GeoStatsDevTools.ZeroMeanIsoNormalCanon

GeoStatsDevTools.binaryentropy

GeoStatsDevTools.boundgrid — Function.

boundgrid(spatialdata, dims)

Returns a RegularGrid of given dims covering all the locations in spatialdata.

GeoStatsDevTools.canonform

GeoStatsDevTools.ccdf

GeoStatsDevTools.cdf

GeoStatsDevTools.cf

GeoStatsDevTools.cgf

GeoStatsDevTools.circvar

GeoStatsDevTools.component

GeoStatsDevTools.components

GeoStatsDevTools.componentwise_logpdf

GeoStatsDevTools.componentwise_pdf

GeoStatsDevTools.concentration

GeoStatsDevTools.cor

GeoStatsDevTools.cov

GeoStatsDevTools.cquantile

GeoStatsDevTools.cumulant

GeoStatsDevTools.dim

GeoStatsDevTools.dof

GeoStatsDevTools.entropy

GeoStatsDevTools.estimate

GeoStatsDevTools.eval

GeoStatsDevTools.evaluate

GeoStatsDevTools.expected_logdet

GeoStatsDevTools.extract_domain

GeoStatsDevTools.failprob

GeoStatsDevTools.fit

GeoStatsDevTools.fit_map

GeoStatsDevTools.fit_map!

GeoStatsDevTools.fit_mle

GeoStatsDevTools.fit_mle!

GeoStatsDevTools.freecumulant

GeoStatsDevTools.gradlogpdf

GeoStatsDevTools.hasfinitesupport

GeoStatsDevTools.include

GeoStatsDevTools.insupport

GeoStatsDevTools.invcov

GeoStatsDevTools.invlogccdf

GeoStatsDevTools.invlogcdf

GeoStatsDevTools.invscale

GeoStatsDevTools.isbounded

GeoStatsDevTools.isleptokurtic

GeoStatsDevTools.islowerbounded

GeoStatsDevTools.ismesokurtic

GeoStatsDevTools.isneighbor — Function.

isneighbor(neighborhood, center, location)

Tells whether or not the location is in the neighborhood centered at center.

isneighbor(neighborhood, xₒ, x)

Tells whether or not the coordinates x is in the neighborhood centered at coordinates xₒ.

Notes

This method is useful in loops, in which case the coordinates can be pre-allocated for better performance.

GeoStatsDevTools.isplatykurtic

GeoStatsDevTools.isprobvec

GeoStatsDevTools.isupperbounded

GeoStatsDevTools.kde

GeoStatsDevTools.kurtosis

GeoStatsDevTools.location

GeoStatsDevTools.location!

GeoStatsDevTools.logccdf

GeoStatsDevTools.logcdf

GeoStatsDevTools.logdetcov

GeoStatsDevTools.loglikelihood

GeoStatsDevTools.logpdf

GeoStatsDevTools.logpdf!

Statistics.mean — Function.

mean(solution)

Mean of simulation solution.

GeoStatsDevTools.meandir

GeoStatsDevTools.meanform

GeoStatsDevTools.meanlogx

GeoStatsDevTools.median

GeoStatsDevTools.mgf

GeoStatsDevTools.mode

GeoStatsDevTools.modes

GeoStatsDevTools.moment

GeoStatsDevTools.ncategories

GeoStatsDevTools.ncomponents

GeoStatsDevTools.nsamples

GeoStatsDevTools.ntrials

GeoStatsDevTools.origin

GeoStatsDevTools.params

GeoStatsDevTools.params!

GeoStatsDevTools.partition — Function.

partition(spatialdata, partitioner)

Partition spatialdata with partition method partitioner.

GeoStatsDevTools.partype

GeoStatsDevTools.pdf

GeoStatsDevTools.probs

GeoStatsDevTools.probval

GeoStatsDevTools.qqbuild

Statistics.quantile — Function.

quantile(solution, p)

p-quantile of simulation solution.

GeoStatsDevTools.rate

GeoStatsDevTools.readgeotable — Function.

readgeotable(args; coordnames=[:x,:y,:z], kwargs)

Read data from disk using CSV.read, optionally specifying the columns coordnames with spatial coordinates.

The arguments args and keyword arguments kwargs are forwarded to the CSV.read function, please check their documentation for more details.

This function returns a GeoDataFrame object.

GeoStatsDevTools.sample

GeoStatsDevTools.sample!

GeoStatsDevTools.sampler

GeoStatsDevTools.scale

GeoStatsDevTools.scale!

GeoStatsDevTools.search! — Function.

search!(neighbors, xₒ, searcher, mask)

Update neighbors of coordinates xₒ with the searcher and mask, and return number of neighbors found.

GeoStatsDevTools.shape

GeoStatsDevTools.skewness

GeoStatsDevTools.spacing

GeoStatsDevTools.span

GeoStatsDevTools.sqmahal

GeoStatsDevTools.sqmahal!

GeoStatsDevTools.std

GeoStatsDevTools.stdlogx

GeoStatsDevTools.subsets — Function.

subsets(partition)

Return the subsets of indices in spatial data that make up the partition.

GeoStatsDevTools.succprob

GeoStatsDevTools.suffstats

GeoStatsDevTools.support

GeoStatsDevTools.test_distr

GeoStatsDevTools.test_samples

GeoStatsDevTools.transform! — Function.

transform!(samples, origin, target)

Transform samples from origin distribution to target distribution in place.

Statistics.var — Function.

var(solution)

Variance of simulation solution.

GeoStatsDevTools.varlogx

GeoStatsDevTools.wsample

GeoStatsDevTools.wsample!