Docstrings

PGFPlotsX.@define_axislike

PGFPlotsX.@pgf — Macro.

@pgf { ... }

@pgf some(nested(form({ ... })),
          with_multiple_options({ ... }))

Construct Options from comma-delimited key (without value), key = value, key : value, or key => value pairs enclosed in { ... }, anywhere in the expression.

The argument is traversed recursively, allowing { ... } expressions in multiple places.

Multi-word keys need to be either quoted, or written with underscores replacing spaces.

@pgf {
    "only marks",
    mark_size = "0.6pt",
    mark = "o",
    color => "black",
}

Another Options can be spliced into one being created using ..., e.g.

theme = @pgf {xmajorgrids, x_grid_style = "white"}

axis_opt = @pgf {theme..., title = "My figure"}

Use {} for empty options that print as [] in LaTeX.

PGFPlotsX.ACTIVE_LATEX_ENGINE

PGFPlotsX.AbstractDict

PGFPlotsX.Axis — Type.

Axis([options], elements...)

Linear axes, corresponds to axis in PGFPlots.

source

PGFPlotsX.AxisElement — Type.

An AxisElement is a component of an Axis. It can be a Plot or a RawString etc.

PGFPlotsX.AxisLike — Type.

abstract type AxisLike <: PGFPlotsX.TikzElement

An axis-like object that has options and contents. Each subtype T has the constructor

T([options], contents...)

and supports axislike_environment(T).

contents usually consists of Plot objects, but can also contain strings, which are printed as is (use these for legends etc). Some subtypes have special semantics, see their documentation.

PGFPlotsX.CLASS_OPTIONS — Constant.

List of class options used in the preamble (default ["tikz"]).

By setting PGFPlotsX.CLASS_OPTIONS[1] = "varwidth"; push!(PGFPlotsX.CLASS_OPTIONS, "crop=false") the preamble will contain documentclass[varwidth,crop=false]{standalone}.

See https://www.ctan.org/pkg/standalone for a list of options.

PGFPlotsX.CUSTOM_FLAGS — Constant.

Custom flags to the engine can be used in the latex command by push!-ing them into the global variable CUSTOM_FLAGS.

PGFPlotsX.CUSTOM_PREAMBLE — Constant.

A vector of stings, added after DEFAULT_PREAMBLE.

Use this for additional definitions \usepackage statements required by the LaTeX code you include into plots.

PGFPlotsX.CUSTOM_PREAMBLE_PATH — Constant.

A file which is spliced directly to the preamble. Customize the file at this path for site-specific setting that apply for every plot.

PGFPlotsX.Colors

PGFPlotsX.Contour

PGFPlotsX.Coordinate — Type.

Coordinate(data; error, errorplus, errorminus, meta)

Construct a coordinate, with optional error bars and metadata. data should be a 2- or 3-element tuples of finite real numbers.

You can specify either

error, which will then be used for error bars in both directions, or

errorplus and/or errorminus, for asymmetrical error bars.

Error values can be tuples of the same kind as data, or nothing.

Metadata can be provided in meta.

Users rarely need to use this constructor, see methods of Coordinates for constructing coordinates from arrays.

Coordinate(x, y; args...)

Convenience constructor for 2-dimensional coordinates.

Coordinate(x, y, z; args...)

Convenience constructor for 3-dimensional coordinates.

PGFPlotsX.CoordinateType

PGFPlotsX.Coordinates — Type.

Coordinates(itr)

Convert the argument, which can be any iterable object, to coordinates.

Specifically,

Coordinate and Nothing are passed through as is,

2- or 3-element tuples of finite real numbers or strings are interpreted as coordinates,

(), and tuples with non-finite numbers become nothing (representing empty lines).

The resulting coordinates are checked for dimension consistency.

Examples

The following are equivalent:

Coordinates((x, 1/x) for x in -5:5)
Coordinates(x == 0 ? () : (x, 1/x) for x in -5:5)
Coordinates(x == 0 ? nothing : Coordinate((x, 1/x)) for x in -5:5)

Use enumerate to add 1, 2, … for the x-axis to an existing set of y coordinates:

Coordinates(enumerate([1, 4, 9]))

Coordinates(x, y; xerror, yerror, xerrorplus, yerrorplus, xerrorminus, yerrorminus, meta)

Two dimensional coordinates from two vectors, with error bars.

Coordinates(x, y, z; xerror, yerror, zerror, xerrorplus, yerrorplus, zerrorplus, xerrorminus, yerrorminus, zerrorminus, meta)

Three dimensional coordinates from two vectors, with error bars.

Coordinates(x, y, z; meta)

Construct coordinates from a matrix of values and edge vectors, such that z[i,j] corresponds to x[i] and y[j]. Empty scanlines are inserted, consistently with the mesh/ordering=x varies option of PGFPlots (the default).

x = range(0; stop = 1, length = 10)
y = range(-1; stop = 2, length = 13)
z = sin.(x) + cos.(y')
Coordinates(x, y, z)

PGFPlotsX.DEBUG

PGFPlotsX.DEFAULT_ENGINE

PGFPlotsX.DEFAULT_FLAGS

PGFPlotsX.DEFAULT_PREAMBLE — Constant.

The default preamble for LaTeX documents. Don't change this, customize CUSTOM_PREAMBLE instead.

PGFPlotsX.DataFrames

PGFPlotsX.Expression — Type.

Expression(expressions::Vector{String})

Expression(strings::String...)

An Expression is a string or multiple strings, representing a function, and is written in a way LaTeX understands.

PGFPlotsX.Graphics — Type.

Graphics([options], filename)

Graphics data simply wraps an image (eg a .png file).

PGFPlotsX.GroupPlot — Type.

GroupPlot([options], contents...)

A group plot, using the groupplots library of PGFPlots.

The contents after the global options are processed as follows:

Options (ie from @pgf {}) will emit a \nextgroupplot with the given options,

nothing is emitted as a \nextgroupplot[group/empty plot],

an Axis is emitted as a \nextgroupplot[options...], followed by the contents,

other values, eg Plot are emitted using print_tex.

source

PGFPlotsX.HAVE_PDFTOPPM

PGFPlotsX.HAVE_PDFTOSVG

PGFPlotsX.HLine — Type.

HLine([options], y)

A horizontal vertical line at y.

PGFPlotsX.Ijulia_cache

PGFPlotsX.LUALATEX

PGFPlotsX.LaTeXEngine

PGFPlotsX.Legend — Type.

Legend(labels)

Corresponds to \legend{ ... } in PGFPlots. Specifies multiple legends for an axis, its position is irrelevant.

labels are wrapped in {}s, so they can contain ,.

PGFPlotsX.LegendEntry — Type.

LegendEntry([options::Options], name, [isexpanded])

Corresponds to the \addlegendentry and \addlegendentryexpanded forms of PGFPlots.

PGFPlotsX.LogLogAxis — Type.

LogLogAxis([options], elements...)

Log-log axes, corresponds to loglogaxis in PGFPlots.

source

PGFPlotsX.MissingExternalProgramError

PGFPlotsX.OptionType — Type.

Subtypes have an options::Options field.

PGFPlotsX.Options — Type.

Options(pairs; print_empty)

Options passed to PGFPlots for various structures (table, plot, etc).

Contents emitted in key = value form, or key when value ≡ nothing. Also see the @pgf convenience macro.

When print_empty = false (the default), empty options are not printed. Use print_empty = true to force printing a [] in this case.

PGFPlotsX.PDFLATEX

PGFPlotsX.PGFOption — Constant.

Types also accepted as options.

PGFPlotsX.PGFPlotsX

PGFPlotsX.PGFPlotsXDisplay

PGFPlotsX.Plot — Type.

struct Plot <: PGFPlotsX.OptionType

Corresponds to the \addplot[3][+] family of pgfplot commands.

Instead of the default constructor, use Plot([options], data, trailing...) and similar (PlotInc, Plot3, Plot3Inc) in user code.

PGFPlotsX.Plot3 — Function.

Plot3([options::Options], data, trailing...)

Corresponds to the \addplot3 form in PGFPlots.

For the interpretation of the other arguments, see Plot(::Options, ::PlotData, ...).

PGFPlotsX.Plot3Inc — Function.

Plot3Inc([options::Options], data, trailing...)

Corresponds to the \addplot3+ form in PGFPlots.

For the interpretation of the other arguments, see Plot(::Options, ::PlotData, ...).

PGFPlotsX.PlotData — Constant.

Types accepted by Plot for the field data.

PGFPlotsX.PlotInc — Function.

PlotInc([options::Options], data, trailing...)

Corresponds to the \addplot+ form in PGFPlots.

For the interpretation of the other arguments, see Plot(::Options, ::PlotData, ...).

PGFPlotsX.PolarAxis — Type.

PolarAxis([options], elements...)

Polar axes, corresponds to polaraxis in PGFPlots.

source

PGFPlotsX.ROWSEP — Constant.

Default for additional row separator \.

PGFPlotsX.STANDALONE_TIKZ_FILEEXTS — Constant.

Extensions that make save choose a standalone tikz format.

The saved file has no preamble, just a tikzpicture environment. These extensions should be recognized by \includegraphics when the tikzscale LaTeX package is used.

PGFPlotsX.SemiLogXAxis — Type.

SemiLogXAxis([options], elements...)

Log x and linear y axes, corresponds to semilogxaxis in PGFPlots.

source

PGFPlotsX.SemiLogYAxis — Type.

SemiLogYAxis([options], elements...)

Linear x and log y axes, corresponds to semilogyaxis in PGFPlots.

source

PGFPlotsX.StatsBase

PGFPlotsX.Table — Type.

Table([options], ...; ...)

Tabular data with options, corresponding to table[options] { ... } in PGFPlots.

options stores the options. If that is followed by an AbstractString, that will be used as a filename to read data from, otherwise all the arguments are passed on to TableData.

Examples:

Table(["x" => 1:10, "y" => 11:20])        # from a vector

Table([1:10, 11:20])                      # same contents, unnamed

Table(Dict(:x => 1:10, :y = 11:20))       # a Dict with symbols

@pgf Table({ "x index" = 2, "y index" = 1" }, randn(10, 3))

let x = range(0; stop = 1, length = 10), y = range(-2; stop =  3, length = 15)
    Table(x, y, sin.(x + y'))             # edges & matrix
end

PGFPlotsX.TableData — Type.

Tabular data with optional column names.

This corresponds to the part of tables between {}'s in PGFPlots, without the options or table, so that it can also be used for “inline” tables. Table will call the constructor for this type to convert arguments after options.

data is a matrix, which contains the contents of the table, which will be printed using print_tex. colnames is a vector of column names (converted to string), or nothing for a table with no column names.

When rowsep is true, an additional \\ is used as a row separator. The default is true, this is recommended to avoid “fragility” issues with inline tables.

Note

Table queries TableData for its rowsep, and adds the relevant option accordingly. When using “inline” tables, eg in options, you have to specify this manually for the container. See the gallery for examples.

After each index in scanlines, extra row separators are inserted. This can be used for skipping coordinates or implicitly defining the dimensions of a matrix for surf and mesh plots. They are expanded using expand_scanlines.

PGFPlotsX.TikzDocument — Type.

TikzDocument(elements...; use_default_preamble = true, preamble = [])

Corresponds to a LaTeX document, usually wrapping TikzPictures.

use_default_preamble determines whether a preamble is added from the global variables (see CUSTOM_PREAMBLE and CUSTOM_PREAMBLE_PATH).

preamble is appended after the default one (if any).

push! can be used to append elements after construction, and similarly push_preamble! for the preamble.

PGFPlotsX.TikzElement — Type.

A TikzElement is a component of a TikzPicture. It can be a node or an Axis etc.

PGFPlotsX.TikzElementOrStr

PGFPlotsX.TikzPicture — Type.

TikzPicture([options], contents...)

Corredponds to a tikzpicture block in PGFPlots.

Elements can also be added with push! after contruction.

PGFPlotsX.VLine — Type.

VLine([options], x)

A vertical line at x.

PGFPlotsX._DISPLAY_PDF

PGFPlotsX._HAS_WARNED_SHELL_ESCAPE

PGFPlotsX._OLD_LUALATEX

PGFPlotsX._SHOWABLE

PGFPlotsX.__init__

PGFPlotsX.__is_juno

PGFPlotsX._default_preamble

PGFPlotsX._engine_cmd

PGFPlotsX._is_ide

PGFPlotsX._is_ijulia

PGFPlotsX._is_juno

PGFPlotsX._is_vscode

PGFPlotsX._replace_fileext — Function.

_replace_fileext(filename, ext)

Replace the extension in filename by ext (which should include the .). When the resulting filename is unchanged, throw an error.

PGFPlotsX._tikzid

PGFPlotsX.accum_opt!

PGFPlotsX.add_indent — Function.

add_indent(str)

Add 4 spaces before each line in str. It is recommended (but not required) that the argument is terminated with a newline.

PGFPlotsX.axislike_environment — Function.

axislike_environment(::Type{<: AxisLike})

Return the corresponding LaTeX environment name.

PGFPlotsX.container_options — Function.

container_options(tabledata, ?)

Options to mix in for the container. Currently used for TableData and Table.

PGFPlotsX.coordinate_or_nothing

PGFPlotsX.dictify

PGFPlotsX.enable_interactive

PGFPlotsX.eval

PGFPlotsX.expand_errors

PGFPlotsX.expand_scanlines — Function.

expand_scanlines(n, nrow)

Expand scanlines, which is a vector of scanline positions or an integer for repeated scanlines, into a Vector{Int}.

PGFPlotsX.include

PGFPlotsX.insert_scanlines — Function.

insert_scanlines(coordinates, stride)

Return new coordinates, inserting nothing after every stride lines.

PGFPlotsX.latexengine

PGFPlotsX.latexengine!

PGFPlotsX.latexerrormsg

PGFPlotsX.matrix_xyz — Function.

matrix_xyz(x, y, z)

Given edges x, y for a matrix z, return a tuple of three vectors which contain the matching elements of x, y, and z, respectively. Useful for surf and mesh plots.

PGFPlotsX.pgfsave — Function.

save(filename, td; include_preamble, latex_engine, buildflags, dpi, showing_ide)

Save the argument (either TikzDocument, or some other type which is wrapped in one automatically, eg TikzPicture, Axis, or Plot) to filename, guessing the format from the file extension. Keywords specify options, some specific to some output formats.

pgfsave is an alias which is exported.

PGFPlotsX.print_indent — Function.

print_indent(f, io_main)

Call the f with an IO buffer, capture the output, print it to io_main indended with four spaces.

print_indent(io_main, elt)

Print elt to io with indentation. Shortcut for the function wrapper of print_indent for a single element.

PGFPlotsX.print_opt

PGFPlotsX.print_options — Function.

print_options(io, options; newline)

Print options between []. For each option, the value is printed using print_opt. Unless newline == true (the default), a newline follows the ], otherwise a space.

PGFPlotsX.print_tex — Function.

print_tex(io, elt, [container])

Print elt to io as LaTeX code. The optional third argument allows methods to work differently depending on the container.

This method should indent as if at the top level, containers indent their contents as necessary. See print_indent.

print_tex(io, str)

Print a string as is, terminated with a newline.

Note

This is used as a workaround for LaTeX code that does not have a corresponding type, eg as elements in Axis. raw or LaTeXStrings are useful to avoid piling up backslashes. The newline is added to separate tokens.

print_tex(io, x)

Real numbers are printed as is, except for non-finite representation.

PGFPlotsX.prockey

PGFPlotsX.procmap

PGFPlotsX.push_preamble! — Function.

push_preamble!(td, items)

Works like push!, but places items in the preamble.

PGFPlotsX.rm_tmpfiles

PGFPlotsX.run_latex_once — Function.

succ, log, cmd =

run_latex_once(filename, eng, flags)

Compile filename with LaTeX engine eng using the given flags.

Return the result of success, the contents of the logfile as a string, and the Cmd object that was run (the latter two useful for diagnostics and informative error messages).

Temporary files (.aux, .log) are cleaned up.

!!! NOTE

Changing the working directory is required because of external tools like
`gnuplot`, which don't respect `--output-directory`.

PGFPlotsX.save — Function.

save(filename, td; include_preamble, latex_engine, buildflags, dpi, showing_ide)

Save the argument (either TikzDocument, or some other type which is wrapped in one automatically, eg TikzPicture, Axis, or Plot) to filename, guessing the format from the file extension. Keywords specify options, some specific to some output formats.

pgfsave is an alias which is exported.

PGFPlotsX.savepdf

PGFPlotsX.savepng

PGFPlotsX.savesvg

PGFPlotsX.savetex

PGFPlotsX.showing_Ijulia