Docstrings

POMDPModels.@pack_InvertedPendulum

POMDPModels.@pack_InvertedPendulum!

POMDPModels.@pack_MountainCar

POMDPModels.@pack_MountainCar!

POMDPModels.@pack_SimpleGridWorld

POMDPModels.@pack_SimpleGridWorld!

POMDPModels.@pack_TMaze

POMDPModels.@pack_TMaze!

POMDPModels.@pack_TMazeState

POMDPModels.@pack_TMazeState!

POMDPModels.@unpack_InvertedPendulum

POMDPModels.@unpack_MountainCar

POMDPModels.@unpack_SimpleGridWorld

POMDPModels.@unpack_TMaze

POMDPModels.@unpack_TMazeState

POMDPModels.AlwaysFeed

POMDPModels.BabyBeliefUpdater

POMDPModels.BabyPOMDP

POMDPModels.DiscreteDistribution

POMDPModels.DummyHeuristic1DPolicy

POMDPModels.Energize

POMDPModels.FeedWhenCrying

POMDPModels.GWPos

POMDPModels.GridWorld

POMDPModels.GridWorldAction

POMDPModels.GridWorldActionSpace

POMDPModels.GridWorldDistribution

POMDPModels.GridWorldState

POMDPModels.GridWorldStateSpace

POMDPModels.InvertedPendulum

POMDPModels.LDNormalStateDist

POMDPModels.LegacyGridWorld

POMDPModels.LightDark1D — Type.

LightDark1D

A one-dimensional light dark problem. The goal is to be near 0. Observations are noisy measurements of the position.

Model

-3-2-1 0 1 2 3 ...| | | | | | | | ... G S

Here G is the goal. S is the starting location

POMDPModels.LightDark1DActionSpace

POMDPModels.LightDark1DState — Type.

LightDark1DState

Fields

y: position

status: 0 = normal, negative = terminal

POMDPModels.MazeBelief

POMDPModels.MazeOptimal

POMDPModels.MazeUpdater

POMDPModels.MountainCar

POMDPModels.POMDPModels

POMDPModels.RandomMDP

POMDPModels.RandomPOMDP

POMDPModels.SimpleGridWorld — Type.

SimpleGridWorld(;kwargs...)

Create a simple grid world MDP. Options are specified with keyword arguments.

States and Actions

The states are represented by 2-element static vectors of integers. Typically any Julia AbstractVector e.g. [x,y] can also be used for arguments. Actions are the symbols :up, :left, :down, and :right.

Keyword Arguments

size::Tuple{Int, Int}: Number of cells in the x and y direction [default: (10,10)]

rewards::Dict: Dictionary mapping cells to the reward in that cell, e.g. Dict([1,2]=>10.0). Default reward for unlisted cells is 0.0

terminate_from::Set: Set of cells from which the problem will terminate. Note that these states are not themselves terminal, but from these states, the next transition will be to a terminal state. [default: Set(keys(rewards))]

tprob::Float64: Probability of a successful transition in the direction specified by the action. The remaining probability is divided between the other neighbors. [default: 0.7]

discount::Float64: Discount factor [default: 0.95]

source

POMDPModels.SmartHeuristic1DPolicy

POMDPModels.Starve

POMDPModels.TIGER_LEFT

POMDPModels.TIGER_LISTEN

POMDPModels.TIGER_OPEN_LEFT

POMDPModels.TIGER_OPEN_RIGHT

POMDPModels.TIGER_RIGHT

POMDPModels.TMaze

POMDPModels.TMazeInit

POMDPModels.TMazeObservationDistribution

POMDPModels.TMazeSpace

POMDPModels.TMazeState

POMDPModels.TMazeStateDistribution

POMDPModels.TMazeStateSpace

POMDPModels.TabularMDP

POMDPModels.TabularPOMDP

POMDPModels.TabularProblem

POMDPModels.TigerActionSpace

POMDPModels.TigerBeliefUpdater

POMDPModels.TigerDistribution

POMDPModels.TigerObservationSpace

POMDPModels.TigerPOMDP

POMDPModels.TigerStateSpace

POMDPModels.a2int

POMDPModels.aind

POMDPModels.cell_ctx

POMDPModels.create_observation_distribution

POMDPModels.create_transition_distribution

POMDPModels.default_sigma

POMDPModels.dir

POMDPModels.dwdt

POMDPModels.euler

POMDPModels.eval

POMDPModels.fill_probability!

POMDPModels.inbounds — Function.

inbounds(mdp::LegacyGridWorld, s::GridWorldState, a::Symbol)

Return false if a is trying to go out of bounds, true otherwise.

POMDPModels.include

POMDPModels.initial_belief

POMDPModels.int2a

POMDPModels.iterator

POMDPModels.posequal

POMDPModels.rk45

POMDPModels.s2i

POMDPModels.sampletype

POMDPModels.static_reward — Function.

static_reward(mdp::LegacyGridWorld, state::GridWorldState)

Return the reward for being in the state (the reward not including bumping)

POMDPModels.tocolor

POMDPModels.upperbound