Docstrings

GeoStatsBase.@estimsolver — Macro.

@estimsolver solver body

A helper macro to create a estimation solver named solver with parameters specified in body. For examples, please check the documentation for @metasolver.

GeoStatsBase.@metasolver — Macro.

@metasolver solver solvertype body

A helper macro to create a solver named solver of type solvertype with parameters specified in body.

Examples

Create a solver with parameters mean and variogram for each variable of the problem, and a global parameter that specifies whether or not to use the GPU:

julia> @metasolver MySolver AbstractSimulationSolver begin
  @param mean = 0.
  @param variogram = GaussianVariogram()
  @global gpu = false
end

Notes

This macro is not intended to be used directly, see other macros defined below for estimation and simulation solvers.

GeoStatsBase.@simsolver — Macro.

@estimsolver solver body

A helper macro to create a simulation solver named solver with parameters specified in body. For examples, please check the documentation for @metasolver.

GeoStatsBase.AbstractDomain — Type.

AbstractDomain{T,N}

A spatial domain with N dimensions in which points are represented with coordinates of type T.

GeoStatsBase.AbstractEstimSolverComparison — Type.

AbstractEstimSolverComparison

A method for comparing estimation solvers.

GeoStatsBase.AbstractEstimationSolver — Type.

AbstractEstimationSolver

A solver for a geostatistical estimation problem.

GeoStatsBase.AbstractMapper — Type.

AbstractMapper

A mapping strategy for mapping spatial data onto a domain.

GeoStatsBase.AbstractProblem — Type.

AbstractProblem

A generic problem in geostatistics.

GeoStatsBase.AbstractSimSolverComparison — Type.

AbstractSimSolverComparison

A method for comparing simulation solvers.

GeoStatsBase.AbstractSimulationSolver — Type.

AbstractSimulationSolver

A solver for a geostatistical simulation problem.

GeoStatsBase.AbstractSolution — Type.

AbstractSolution

A generic solution to a problem in geostatistics.

GeoStatsBase.AbstractSolver — Type.

AbstractSolver

A solver for geostatistical problems.

GeoStatsBase.AbstractSolverComparison — Type.

AbstractSolverComparison

A method for comparing geostatistical solvers.

GeoStatsBase.AbstractSpatialData — Type.

AbstractSpatialData{T,N}

Spatial data distributed in a N-dimensional space using coordinates of type T.

GeoStatsBase.CopyMapper — Type.

CopyMapper

A mapping strategy in which data points are copied directly to the domain at specified locations.

GeoStatsBase.DomainView — Type.

DomainView(domain, locations)

Return a view of domain at locations.

Notes

This type implements the AbstractDomain interface.

GeoStatsBase.EstimationProblem — Type.

EstimationProblem(spatialdata, domain, targetvars)

A spatial estimation problem on a given domain in which the variables to be estimated are listed in targetvars. The data of the problem is stored in spatialdata.

Examples

Create an estimation problem for rainfall precipitation measurements:

julia> EstimationProblem(spatialdata, domain, :precipitation)

Create an estimation problem for precipitation and CO₂:

julia> EstimationProblem(spatialdata, domain, [:precipitation, :CO₂])

GeoStatsBase.EstimationSolution — Type.

EstimationSolution

A solution to a spatial estimation problem.

GeoStatsBase.GeoStatsBase

GeoStatsBase.SimpleMapper — Type.

SimpleMapper

A mapping strategy in which data points are assigned to their nearest point in the domain.

GeoStatsBase.SimulationProblem — Type.

SimulationProblem(spatialdata, domain, targetvars, nreals)
SimulationProblem(domain, targetvars, nreals)

A spatial simulation problem on a given domain in which the variables to be simulated are listed in targetvars.

For conditional simulation, the data of the problem is stored in spatialdata.

For unconditional simulation, a dictionary targetvars must be provided mapping variable names to their types.

In both cases, a number nreals of realizations is requested.

Examples

Create a conditional simulation problem for porosity and permeability with 100 realizations:

julia> SimulationProblem(spatialdata, domain, [:porosity,:permeability], 100)

Create an unconditional simulation problem for porosity and facies type with 100 realizations:

julia> SimulationProblem(domain, Dict(:porosity => Float64, :facies => Int), 100)

Notes

To check if a simulation problem has data (i.e. conditional vs. unconditional) use the hasdata method.

GeoStatsBase.SimulationSolution — Type.

SimulationSolution

A solution to a spatial simulation problem.

GeoStatsBase.SpatialDataView — Type.

SpatialDataView(spatialdata, inds)

Return a view of spatialdata at inds.

Notes

This type implements the AbstractSpatialData interface.

GeoStatsBase.compare — Function.

compare(solvers, problem, method)

Compare solvers on a given problem using a comparison method.

GeoStatsBase.coordinates — Function.

coordinates(spatialdata, ind)

Return the coordinates of the ind-th point in spatialdata.

coordinates(spatialdata, inds)

Return the coordinates of inds in the spatialdata.

coordinates(spatialdata)

Return the coordinates of all indices in spatialdata.

coordinates(domain, location)

Return the coordinates of the location in the domain.

coordinates(domain, locations)

Return the coordinates of locations in the domain.

coordinates(domain)

Return the coordinates of all locations in domain.

GeoStatsBase.coordinates! — Function.

coordinates!(buff, spatialdata, indices)

Non-allocating version of coordinates

coordinates!(buff, spatialdata, ind)

Non-allocating version of coordinates.

coordinates!(buff, domain, locations)

Non-allocating version of coordinates

coordinates!(buff, domain, location)

Non-allocating version of coordinates.

GeoStatsBase.coordnames — Function.

coordnames(spatialdata)

Return the name of the coordinates in spatialdata.

GeoStatsBase.coordtype — Function.

coordtype(spatialdata)

Return the coordinate type of a spatial data.

coordtype(domain)

Return the coordinate type of a spatial domain.

GeoStatsBase.data — Function.

data(problem)

Return the spatial data of the estimation problem.

data(problem)

Return the spatial data of the simulation problem.

GeoStatsBase.datamap — Function.

datamap(problem, targetvar)

Return the mapping from domain locations to data locations for the targetvar of the problem.

datamap(problem)

Return the mappings from domain locations to data locations for all the variables of the problem.

datamap(problem, targetvar)

Return the mapping from domain locations to data locations for the targetvar of the problem.

datamap(problem)

Return the mappings from domain locations to data locations for all the variables of the problem.

GeoStatsBase.digest — Function.

digest(solution)

Convert solution to a dictionary-like format where the keys of the dictionary are the variables of the problem.

digest(solution)

Convert solution to a dictionary-like format where the keys of the dictionary are the variables of the problem.

GeoStatsBase.domain — Function.

domain(problem)

Return the spatial domain of the estimation problem.

domain(problem)

Return the spatial domain of the simulation problem.

domain(solution)

Return the domain of the estimation solution.

domain(solution)

Return the domain of a simulation solution.

GeoStatsBase.eval

GeoStatsBase.hasdata — Function.

hasdata(problem)

Return true if simulation problem has data.

GeoStatsBase.include

GeoStatsBase.mapper — Function.

mapper(problem)

Return the mapper of the estimation problem.

mapper(problem)

Return the mapper of the simulation problem.

GeoStatsBase.nearestlocation — Function.

nearestlocation(domain, coords)

Return the nearest location of coords in the domain.

GeoStatsBase.npoints — Function.

npoints(spatialdata)

Return the number of points in spatialdata.

npoints(domain)

Return the number of points of a spatial domain.

GeoStatsBase.nreals — Function.

nreals(problem)

Return the number of realizations of the simulation problem.

GeoStatsBase.preprocess — Function.

preprocess(problem, solver)

Preprocess the simulation problem once before generating each realization with simulation solver.

Notes

The output of the function is defined by the solver developer. Default implementation returns nothing.

GeoStatsBase.solve — Function.

solve(problem, solver)

Solve the problem with the solver.

solve(problem, solver)

Solve the simulation problem with the simulation solver.

Notes

Default implementation calls solve_single in parallel.

GeoStatsBase.solve_single — Function.

solve_single(problem, var, solver, preproc)

Solve a single realization of var in the simulation problem with the simulation solver, optionally using preprocessed data in preproc.

Notes

By implementing this function instead of solve, the developer is informing the framework that realizations generated with his/her solver are indenpendent one from another. GeoStats.jl will trigger the algorithm in parallel (if enough processes are available).

GeoStatsBase.valid — Function.

valid(spatialdata, var)

Return all points in spatialdata with a valid value for var. The output is a tuple with the matrix of coordinates as the first item and the vector of values as the second item.

GeoStatsBase.value — Function.

value(spatialdata, ind, var)

Return the value of var for the ind-th point in spatialdata.

GeoStatsBase.valuetype — Function.

valuetype(spatialdata, var)

Return the value type of var in spatialdata.

GeoStatsBase.variables — Function.

variables(spatialdata)

Return the variable names in spatialdata and their types.

variables(problem)

Return the variable names of the estimation problem and their types.

variables(problem)

Return the target variables of the simulation problem and their types.