Docstrings

NFFT.NFFT

NFFT.NFFTPlan — Type.

    NFFTPlan(x, N, ...) -> plan

Compute D dimensional NFFT plan for sampling locations x (a vector or a D-by-M matrix) that can be applied on arrays of size N (a tuple of length D).

The optional arguments control the accuracy.

It takes as optional keywords all the keywords supported by plan_fft function (like flags and timelimit). See documentation of plan_fft for reference.

NFFT._convolve

NFFT.apodization!

NFFT.apodization_adjoint!

NFFT.cbspline

NFFT.consistencyCheck

NFFT.convolve!

NFFT.convolve_adjoint!

NFFT.dim

NFFT.eval

NFFT.getWindow

NFFT.include

NFFT.ndft

NFFT.ndft_adjoint

NFFT.nfft — Function.

    nfft(p, f) -> fHat

For a non-directional D dimensional plan p this calculates the NFFT of a D dimensional array f of size N. fHat is a vector of length M. (M and N are defined in the plan creation)

For a directional D dimensional plan p both f and fHat are D dimensional arrays, and the dimension specified in the plan creation is affected.

NFFT.nfft! — Function.

    nfft!(p, f, fHat) -> fHat

Calculate the NFFT of f with plan p and store the result in fHat.

Both f and fHat must be complex arrays.

NFFT.nfft_adjoint — Function.

    nfft_adjoint(p, f) -> fHat

For a non-directional D dimensional plan p this calculates the adjoint NFFT of a length M vector fHat f is a D dimensional array of size N. (M and N are defined in the plan creation)

For a directional D dimensional plan p both f and fHat are D dimensional arrays, and the dimension specified in the plan creation is affected.

NFFT.nfft_adjoint! — Function.

    nfft_adjoint!(p, fHat, f) -> f

Calculate the adjoint NFFT of fHat and store the result in f.

Both f and fHat must be complex arrays.

NFFT.sdc

NFFT.window_gauss

NFFT.window_gauss_hat

NFFT.window_kaiser_bessel

NFFT.window_kaiser_bessel_hat

NFFT.window_kaiser_bessel_rev

NFFT.window_kaiser_bessel_rev_hat

NFFT.window_spline

NFFT.window_spline_hat