Docstrings

ZChop.ZChop — Module.

The ZChop module provides the functions zchop and zchop!, which replace numbers that are close to zero with zero. These functions act recursivley on collections.

ZChop.eval

ZChop.include

ZChop.zchop — Function.

zchop(x::T, eps::Real = zeps)

Replace x by zero if abs(x) < eps. zchop acts recursively on mappable objects. Objects that cannot be sensibly compared to a real number are passed unaltered.

ZChop.zchop! — Function.

zchop!(x::T, eps::Real = zeps)

Perform zchop in place.

ZChop.zeps — Constant.

zeps::Float64

The default lower threshold for a number to be replaced by zero.