Docstrings

Algencan.@checked_lib

Algencan.Algencan — Module.

Algencan interface to MathProgBase and JuMP.

See its GitHub page

Algencan.AlgencanMathProgModel — Type.

Algencan model, that storing solution data

Algencan.AlgencanSolver — Type.

Algencan solver that stores options

Algencan.MPB

Algencan.algencan_lib_path

Algencan.eval

Algencan.find_status — Function.

Find final status of Algencan

Algencan.getnfevals

Algencan.include

Algencan.julia_fc — Function.

Compute objective and constraints as required by Algencan

Algencan.julia_gjac — Function.

Compute objective gradient and constraints Jacobian as required by Algencan

Algencan.julia_hl — Function.

Compute the Hessian of the Lagrangian as required by Algencan

Algencan.julia_hlp — Function.

Compute the Hessian of the Lagrangian times p as required by Algencan

Algencan.libalgencan

Algencan.option2vparam — Function.

Transform the option dictionary into a vparam string array

Algencan.resetnfevals

Algencan.setmultwarmstart! — Function.

Set an inital value for the constaint mutipliers (warmstart)

Algencan.treat_lower_bounds — Function.

Analyse the lower and upper bounds on the constraints and prepare the data structure to treat lower bounds.