Docstrings

Phylo.API — Module.

Phylo.API submodule

The Phylo.API submodule should be imported if you want to create a new phylogeny, node or branch subtype. Otherwise it can be ignored.

Phylo.AbstractBranch

Phylo.AbstractBranchIterator

Phylo.AbstractBranchTree

Phylo.AbstractNode

Phylo.AbstractNodeIterator

Phylo.AbstractTree

Phylo.AbstractTreeIterator

Phylo.BinaryNode — Type.

BinaryNode{T}(AbstractVector{T}, AbstractVector{T}) <: AbstractNode

A node of strict binary phylogenetic tree

Phylo.BinaryTree — Type.

BinaryTree

Binary phylogenetic tree object with known leaves and per node data

Phylo.Branch — Type.

Branch

A directed branch connecting two AbstractNodes of phylogenetic tree

Phylo.BranchIterator

Phylo.BranchNameIterator

Phylo.Dendrogram

Phylo.Fan

Phylo.LeafInfo

Phylo.NamedBinaryTree — Type.

BinaryTree

Binary phylogenetic tree object with known leaves and per node data

Phylo.NamedPolytomousTree — Type.

NamedPolytomousTree

Polytomous phylogenetic tree object with known leaves

Phylo.NamedTree — Type.

NamedBinaryTree

Binary phylogenetic tree object with known leaves

Phylo.Node — Type.

Node{T}(AbstractVector{T}, AbstractVector{T}) <: AbstractNode

A node of potentially polytomous phylogenetic tree

Phylo.NodeIterator

Phylo.NodeNameIterator

Phylo.Nonultrametric — Type.

Nonultrametric{T <: AbstractTree,
               RNG <: Sampleable}(n::Int,
                                  rng::RNG = Exponential())
Nonultrametric{T <: AbstractTree,
               RNG <: Sampleable}(tiplabels::Vector{String},
                                  rng::RNG = Exponential())

The sampler for non-ultrametric phylogenetic trees of size n or with tip labels tiplabels. Generate random trees by calling rand(). Currently only works for NamedTrees.

Phylo.Phylo — Module.

Phylo package

The Phylo package provides some simple phylogenetics types (e.g. NamedTree) to interface to the Diversity package for measuring phylogenetic diversity. It also provides an interface to R for copying trees to and from that language and can read newick and nexus tree files (including TreeSets that contain multiple trees).

Finally it also provides a standard abstract interface to phylogenetic trees, by defining AbstractNode, AbstractBranch and AbstractTree supertypes, and methods to interface to them. It also provides (through the Phylo.API submodule) methods to (re)define to write your own phylogenetic type in a way that will interact cleanly with other phylogenetic packages.

Phylo.Phylogenetics

Phylo.PolytomousTree — Type.

PolytomousTree

Phylogenetic tree object with polytomous branching, and known leaves and per node data

Phylo.SimpleBranch — Type.

Branch

A directed branch connecting two AbstractNodes of phylogenetic tree

Phylo.T

Phylo.TreeInfoIterator

Phylo.TreeIterator

Phylo.TreeNameIterator

Phylo.TreeSet

Phylo.Ultrametric — Type.

Ultrametric{T <: AbstractTree,
            RNG <: Sampleable}(n::Int,
                               rng::RNG = Exponential())
Ultrametric{T <: AbstractTree,
            RNG <: Sampleable}(tiplabels::Vector{String},
                               rng::RNG = Exponential())

The sampler for ultrametric phylogenetic trees of size n or with tip labels tiplabels. Generate random trees by calling rand(). Currently only works for NamedTrees.

Phylo.__init__

Phylo._circle_transform_segments

Phylo._extend

Phylo._find_tips

Phylo._findxy

Phylo._hasbranch

Phylo._hasnode

Phylo._p_circ

Phylo._tocirc

Phylo._treefuture

Phylo._treepast

Phylo._xcirc

Phylo._ycirc

Phylo.addbranch! — Function.

addbranch!(tree::AbstractTree, source, destination[, length::Float64];
           branchname = _newbranchlabel(tree))

Add a branch from source to destination on tree.

Phylo.addnode! — Function.

addnode!(tree::AbstractTree)
addnode!(tree::AbstractTree, nodename)

Phylo.addnodes! — Function.

addnodes!(tree::AbstractTree, nodenames::AbstractVector)
addnodes!(tree::AbstractTree, count::Integer)

Phylo.branch! — Function.

branch!(tree::AbstractTree, source[, length])
branch!(tree::AbstractTree, source[, length]; destination)
branch!(tree::AbstractTree, source[, length]; destination, branchname)

Branch from a source node source and create a destination node destination.

Phylo.branchfilter — Function.

branchfilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the branches of any tree, where the AbstractBranch is filtered by the function filterfn.

Phylo.branchhistory — Function.

branchhistory(tree::AbstractTree, node)

Find the branch route between a node on a tree and its root

Phylo.branchiter — Function.

branchiter(tree::AbstractTree)

Returns an iterator over the branches of any tree.

Phylo.branchnamefilter — Function.

branchnamefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the names of the branches of any tree, where the AbstractBranch is filtered by the function filterfn.

Phylo.branchnameiter — Function.

branchnameiter(tree::AbstractTree)

Returns an iterator over the names of branches of any tree.

Phylo.branchnametype — Function.

branchnametype(::AbstractTree)

Returns type of branch names.

Phylo.branchroute — Function.

branchroute(tree::AbstractTree, node1, node2)

Find the branch route between two nodes on a tree

Phylo.branchtype — Function.

branchtype(tree::AbstractTree)

Returns type of branches in a tree.

Phylo.changedst! — Function.

changedst!(tree::AbstractTree, branchname, destination)

Change the destination node for this node.

Phylo.changesrc! — Function.

changesrc!(tree::AbstractTree, branchname, source)

Change the source node for this branch.

Phylo.checkbranch

Phylo.checktosemi

Phylo.clearrootheight! — Function.

clearrootheight(::AbstractTree)

Clears the tree's root height record.

Phylo.deletebranch! — Function.

deletebranch!(tree::AbstractTree, branchname)

Delete the branch branchname from tree.

Phylo.deletenode! — Function.

deletenode!(tree::AbstractTree, nodename)

Phylo.distance — Function.

distance(tree::AbstractTree, node1, node2)

Distance between two nodes on a tree

Phylo.distances — Function.

distances(tree::AbstractTree)

Pairwise distances between all leaf nodes on a tree

Phylo.droptips! — Function.

droptips!(t::T, tips::Vector{NL}) where {NL, BL, T <: AbstractTree{NL, BL}}

Function to drop tips from a phylogenetic tree t, which are found in the vector of tip names, tips.

Phylo.dst — Function.

dst(branch::AbstractBranch)
dst(tree::AbstractTree, branchname)

Return the destination node for this branch.

Phylo.eval

Phylo.getancestors — Function.

getancestors(tree::AbstractTree, nodename)

Return the name of all of the nodes that are ancestral to this node.

Phylo.getbranch — Function.

getbranch(tree::AbstractTree, branchname)

Phylo.getbranches

Phylo.getbranchnames — Function.

getbranchnames(tree::AbstractTree)

Phylo.getchildren — Function.

getchildren(tree::AbstractTree, nodename)

Return the name(s) of the child node(s) for this node.

Phylo.getdescendants — Function.

getdescendants(tree::AbstractTree, nodename)

Return the names of all of the nodes that descend from this node.

Phylo.getheight — Function.

getheight(tree::AbstractTree, nodename)

Return the height of the node.

Phylo.getinbound — Function.

getinbound(node::AbstractNode)
getinbound(tree::AbstractTree, nodename)

return the name of the inbound branch to this node.

Phylo.getinternalnodes — Function.

getinternalnodes(t::AbstractTree)

Function to retrieve only the internal nodes from a tree, t, which does not include tips or root.

Phylo.getleafinfo — Function.

getleafinfo(::AbstractTree, label)

retrieve the leaf info for a leaf of the tree.

Phylo.getleafnames — Function.

getleafnames(::AbstractTree)

Retrieve the leaf names from the tree.

Phylo.getlength — Function.

getlength(branch::AbstractBranch)
getlength(tree::AbstractTree, branchname)

Return the length of this branch.

Phylo.getnode — Function.

getnode(tree::AbstractTree, nodename)

Phylo.getnodenames — Function.

getnodenames(tree::AbstractTree)

Phylo.getnoderecord — Function.

getnoderecord(::AbstractTree, label)

retrieve the node record for a leaf of the tree.

Phylo.getnodes

Phylo.getoutbounds — Function.

getoutbounds(node::AbstractNode)
getoutbounds(tree::AbstractTree, nodename)

Return the names of the outbound branches from this node.

Phylo.getparent — Function.

getparent(tree::AbstractTree, nodename)

Return the name of the parent node for this node.

Phylo.getrootdistance

Phylo.getrootheight — Function.

getrootheight(tree::AbstractTree)

Phylo.hasbranch — Function.

hasbranch(tree::AbstractTree, branchname)

Phylo.haschildren

Phylo.hasheight — Function.

hasheight(tree::AbstractTree, nodename)

Does the node have a height defined?

Phylo.hasinbound — Function.

hasinbound(node::AbstractNode)
hasinbound(tree::AbstractTree, nodename)

Does the node have an inbound connection?

Phylo.hasinboundspace — Function.

hasinboundspace(node::AbstractNode)
hasinboundspace(tree::AbstractTree, nodename)

Does the node have space for an inbound connection?

Phylo.hasnode — Function.

hasnode(tree::AbstractTree, nodename)

Phylo.hasoutboundspace — Function.

hasoutboundspace(node::AbstractNode)
hasoutboundspace(tree::AbstractTree, nodename)

Does the node have space for an[other] outbound connection?

Phylo.hasparent

Phylo.hasrootheight — Function.

hasrootheight(tree::AbstractTree)

Phylo.heightstoroot — Function.

heights(tree::AbstractTree)

Height of all of the leaves of the tree above the root

Phylo.heighttoroot — Function.

height(tree::AbstractTree, node)

Height of a node of the tree above the root

Phylo.include

Phylo.indegree — Function.

indegree(node::AbstractNode)
indegree(tree::AbstractTree, nodename)

Phylo.isBEGIN

Phylo.isDIMENSIONS

Phylo.isEND

Phylo.isEQ

Phylo.isEQorRSQUARE

Phylo.isIDENTIFIER

Phylo.isTAXA

Phylo.isTAXLABELS

Phylo.isTRANSLATE

Phylo.isTREE

Phylo.isTREES

Phylo.isWHITESPACE

Phylo.isinternal — Function.

isinternal(node::AbstractNode)
isinternal(tree::AbstractTree, nodename)

Phylo.isleaf — Function.

isleaf(node::AbstractNode)
isleaf(tree::AbstractTree, nodename)

Phylo.isroot — Function.

isroot(node::AbstractNode)
isroot(tree::AbstractTree, nodename)

Phylo.isunattached — Function.

isunattached(node::AbstractNode)
isunattached(tree::AbstractTree, nodename)

Phylo.iterateskip

Phylo.keeptips! — Function.

keeptips!(t::T, tips::Vector{NL}) where {NL, BL, T <: AbstractTree{NL, BL}}

Function to keep only the tips in a phylogenetic tree, t, that are found in the vector of tip names, tip.

Phylo.leafinfotype

Phylo.leafiter

Phylo.map_depthfirst

Phylo.nleaves — Function.

nleaves(::AbstractTree)

Count the number of leaves in the tree.

Phylo.nodefilter — Function.

nodefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the nodes of any tree, where the AbstractNode is filtered by the function filterfn.

Phylo.nodehistory — Function.

nodehistory(tree::AbstractTree, node)

Find the node route between a node on a tree and its root

Phylo.nodeiter — Function.

nodeiter(tree::AbstractTree)

Returns an iterator over the nodes of any tree.

Phylo.nodenamefilter — Function.

nodenamefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the nodenames of any tree, where the AbstractNode itself is filtered by the function filterfn.

Phylo.nodenameiter — Function.

nodenameiter(tree::AbstractTree)

Returns an iterator over the names of the nodes of any tree.

Phylo.nodenametype — Function.

nodenametype(::AbstractTree)

Returns type of node names.

Phylo.noderoute — Function.

noderoute(tree::AbstractTree, node1, node2)

Find the node route between two nodes on a tree

Phylo.nodetype — Function.

nodetype(tree::AbstractTree)

Returns type of nodes in a tree.

Phylo.noname

Phylo.ntrees

Phylo.outdegree — Function.

outdegree(node::AbstractNode)
outdegree(tree::AbstractTree, nodename)

Phylo.parsedict

Phylo.parsenewick

Phylo.parsenewick!

Phylo.parsenexus

Phylo.parsenode

Phylo.parsetaxa

Phylo.parsetrees

Phylo.parsevector

Phylo.path

Phylo.resetleaves

Phylo.resetleaves! — Function.

resetleaves!(::AbstractTree)

Reset the leaf records to the current leaves, deleting all leaf records.

Phylo.setheight! — Function.

setheight!(tree::AbstractTree, nodename, height)

Set the height of the node.

Phylo.setleafinfo! — Function.

setleafinfo!(::AbstractTree, table)

Set the leaf info for the leaves of the tree.

Phylo.setnoderecord! — Function.

setnoderecord(::AbstractTree, label, value)

Set the node record for a node of the tree.

Phylo.setrootheight! — Function.

setrootheight!(tree::AbstractTree, height)

Phylo.showsimple

Phylo.src — Function.

src(branch::AbstractBranch)
src(tree::AbstractTree, branchname)

Return the source node for this branch.

Phylo.tokenerror

Phylo.tokensgetkey

Phylo.treeinfoiter

Phylo.treeiter

Phylo.treenameiter

Phylo.validate — Function.

validate(tree::AbstractTree)