Docstrings

Setfield.@lens — Macro.

@lens

Construct a lens from a field access.

Example

julia> using Setfield

julia> struct T;a;b;end

julia> t = T("A1", T(T("A3", "B3"), "B2"))
T("A1", T(T("A3", "B3"), "B2"))

julia> l = @lens _.b.a.b
(@lens _.b.a.b)

julia> get(t, l)
"B3"

julia> set(t, l, 100)
T("A1", T(T("A3", 100), "B2"))

julia> t = ("one", "two")
("one", "two")

julia> set(t, (@lens _[1]), "1")
("1", "two")

Setfield.@set — Macro.

@set assignment

Return a modified copy of deeply nested objects.

Example

julia> using Setfield

julia> struct T;a;b end

julia> t = T(1,2)
T(1, 2)

julia> @set t.a = 5
T(5, 2)

julia> t
T(1, 2)

julia> t = @set t.a = T(2,2)
T(T(2, 2), 2)

julia> @set t.a.b = 3
T(T(2, 3), 2)

Setfield.@set! — Macro.

@set! assignment

Shortcut for obj = @set obj....

julia> t = (a=1,) (a = 1,)

julia> @set! t.a=2 (a = 2,)

julia> t (a = 2,)

Setfield.@settable

Setfield.ComposedLens

Setfield.ConstIndexLens — Type.

ConstIndexLens{I}

Lens with index stored in type parameter. This is useful for type-stable get and set operations on tuples and named tuples.

This lens can be constructed by, e.g., @lens _[$1]. Complex expression must be wrapped with $(...) like @lens _[$(length(xs))].

Examples

julia> using Setfield

julia> get((1, 2.0), @lens _[$1])
1

julia> Base.promote_op(get, typeof.(((1, 2.0), @lens _[$1]))...)
Int64

julia> Base.promote_op(get, typeof.(((1, 2.0), @lens _[1]))...) !== Int
true

Setfield.Experimental

Setfield.IdentityLens

Setfield.IndexLens

Setfield.Lens — Type.

Lens

A Lens allows to access or replace deeply nested parts of complicated objects.

Example

julia> using Setfield

julia> struct T;a;b; end

julia> obj = T("AA", "BB")
T("AA", "BB")

julia> lens = @lens _.a
(@lens _.a)

julia> get(obj, lens)
"AA"

julia> set(obj, lens, 2)
T(2, "BB")

julia> obj
T("AA", "BB")

julia> modify(lowercase, obj, lens)
T("aa", "BB")

Interface

Concrete subtypes of Lens have to implement

set(obj, lens, val)

get(obj, lens)

These must be pure functions, that satisfy the three lens laws:

@assert get(set(obj, lens, val), lens) == val
        # You get what you set.
@assert set(obj, lens, get(obj, lens)) == obj
        # Setting what was already there changes nothing.
@assert set(set(obj, lens, val1), lens, val2) == set(obj, lens, val2)
        # The last set wins.

See also @lens, set, get, modify.

Setfield.PropertyLens

Setfield.Setfield

Setfield._UpdateOp

Setfield.add_posonly_constructor

Setfield.arg_type

Setfield.argsymbol

Setfield.argsymbol_typed

Setfield.assert_hasfields

Setfield.atset_impl

Setfield.best_constructor_template

Setfield.compose

Setfield.constructor_has_desired_fields

Setfield.constructor_of

Setfield.eval

Base.get — Function.

get(obj, l::Lens)

Access a deeply nested part of obj. See also Lens.

Setfield.get_update_op

Setfield.has_atlens_support

Setfield.has_posonly_constructor

Setfield.include

Setfield.is_interpolation

Setfield.modify — Function.

modify(f, obj, l::Lens)

Replace a deeply nested part x of obj by f(x). See also Lens.

Setfield.parse_obj_lens

Setfield.parse_obj_lenses

Setfield.posonly_constructor

Setfield.posonly_constructor_dict

Setfield.print_application

Setfield.print_in_atlens

Setfield.set — Function.

set(obj, l::Lens, val)

Replace a deeply nested part of obj by val. See also Lens.

Setfield.setproperties

Setfield.settable

Setfield.show_generic