Docstrings

Variography.@pack_CubicVariogram

Variography.@pack_CubicVariogram!

Variography.@pack_ExponentialVariogram

Variography.@pack_ExponentialVariogram!

Variography.@pack_GaussianVariogram

Variography.@pack_GaussianVariogram!

Variography.@pack_MaternVariogram

Variography.@pack_MaternVariogram!

Variography.@pack_PentasphericalVariogram

Variography.@pack_PentasphericalVariogram!

Variography.@pack_PowerVariogram

Variography.@pack_PowerVariogram!

Variography.@pack_SineHoleVariogram

Variography.@pack_SineHoleVariogram!

Variography.@pack_SphericalVariogram

Variography.@pack_SphericalVariogram!

Variography.@unpack_CubicVariogram

Variography.@unpack_ExponentialVariogram

Variography.@unpack_GaussianVariogram

Variography.@unpack_MaternVariogram

Variography.@unpack_PentasphericalVariogram

Variography.@unpack_PowerVariogram

Variography.@unpack_SineHoleVariogram

Variography.@unpack_SphericalVariogram

Variography.CompositeVariogram — Type.

CompositeVariogram(γ₁, γ₂, ..., γₙ)

A composite (additive) model of variograms γ(h) = γ₁(h) + γ₂(h) + ⋯ + γₙ(h).

Variography.CubicVariogram — Type.

CubicVariogram(sill=s, range=r, nugget=n, distance=d)

A cubic variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Variography.DirectionalVariogram — Function.

DirectionalVariogram(spatialdata, direction, var₁, var₂=var₁; [optional parameters])

Computes the empirical (cross-)variogram for the variables var₁ and var₂ stored in spatialdata along a given direction.

Optional parameters include the parameters for EmpiricalVariogram and the parameters for DirectionPartitioner.

Notes

A DirectionalVariogram is just a function that first partitions the spatialdata using a DirectionPartitioner and then passes the result to the corresponding EmpiricalVariogram constructor.

See also: EmpiricalVariogram, DirectionPartitioner

Variography.EmpiricalVariogram — Type.

EmpiricalVariogram(X, z₁, z₂=z₁; [optional parameters])

Computes the empirical (a.k.a. experimental) omnidirectional (cross-)variogram from data locations X and values z₁ and z₂.

EmpiricalVariogram(spatialdata, var₁, var₂=var₁; [optional parameters])

Alternatively, compute the (cross-)variogram for the variables var₁ and var₂ stored in a spatialdata object.

EmpiricalVariogram(partition, var₁, var₂=var₁; [optional parameters])

Alternatively, compute the (cross-)variogram on a partition of the data.

Parameters

nlags - number of lags (default to 20)

maxlag - maximum lag (default to half of maximum lag of data)

distance - custom distance function (default to Euclidean distance)

See also: DirectionalVariogram

Variography.ExponentialVariogram — Type.

ExponentialVariogram(sill=s, range=r, nugget=n, distance=d)

An exponential variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Variography.FitAlgo — Type.

FitAlgo

An algorithm for fitting theoretical variograms.

Variography.GaussianVariogram — Type.

GaussianVariogram(sill=s, range=r, nugget=n, distance=d)

A Gaussian variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Variography.HScatter

Variography.MaternVariogram — Type.

MaternVariogram(sill=s, range=r, nugget=n, order=ν, distance=d)

A Matérn variogram with sill s, range r and nugget n. The parameter ν is the order of the Bessel function. Optionally, use a custom distance d.

source

Variography.PentasphericalVariogram — Type.

PentasphericalVariogram

A pentaspherical variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Variography.PowerVariogram — Type.

PowerVariogram(scaling=s, exponent=a, nugget=n, distance=d)

A power variogram with scaling s, exponent a and nugget n. Optionally, use a custom distance d.

source

Variography.SineHoleVariogram — Type.

SineHoleVariogram(sill=s, range=r, nugget=n, distance=d)

A sine hole variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Variography.SphericalVariogram — Type.

SphericalVariogram(sill=s, range=r, nugget=n, distance=d)

A spherical variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Variography.VarPlane

Variography.Variogram — Type.

Variogram

A variogram model (e.g. Gaussian variogram).

Variography.Variography

Variography.WeightedLeastSquares — Type.

WeightedLeastSquares(weightfun)

Fit theoretical variogram using weighted least squares with weighting function weightfun (e.g. h -> 1/h). If not weighting function is provided, bin counts of empirical variogram are normalized and used as weights.

Variography.eval

Variography.fit — Function.

fit(V, γ, [algo])

Fit theoretical variogram type V to empirical variogram γ using algorithm algo. Default algorithm is WeightedLeastSquares.

fit(Variogram, γ, [algo])

Fit all stationary variogram types to empirical variogram γ, which are subtypes of Variogram, and return the one with minimum error as defined by the algorithm algo.

Variography.fit_impl

Variography.hscatter

Variography.hscatter!

Variography.include

Variography.isstationary — Function.

isstationary(γ)

Check if variogram γ possesses the 2nd-order stationary property.

Variography.merge! — Function.

merge!(γₐ, γᵦ)

Merge the empirical variogram γₐ with the empirical variogram γᵦ assuming that both have the same abscissa.

Variography.pairwise — Function.

pairwise(γ, X)

Evaluate variogram γ between all n² pairs of columns in a m-by-n matrix X efficiently.

pairwise(γ, domain, locations)

Evaluate variogram γ between all locations in domain.

pairwise(γ, domain, locations₁, locations₂)

Evaluate variogram γ between locations₁ and locations₂ in domain.

Variography.pairwise! — Function.

pairwise!(Γ, γ, X)

Non-allocating pairwise evaluation.

Variography.param_type — Function.

param_type(γ)

Return the parameter (e.g. sill, range) type of the variogram.

Variography.plane_basis

Variography.result_type — Function.

result_type(γ, x₁, x₂)

Return result type of γ(x₁, x₂).

Variography.sill — Function.

sill(γ)

Return the sill of the variogram γ when defined.

Variography.varplane

Variography.varplane!