Docstrings

MicroLogging.@debug

MicroLogging.@error

MicroLogging.@info

MicroLogging.@logmsg

MicroLogging.@warn

MicroLogging.AbstractLogger

MicroLogging.ConsoleLogger — Type.

ConsoleLogger(stream=stderr, min_level=Info; meta_formatter=default_metafmt,
              show_limited=true, right_justify=0)

Logger with formatting optimized for readability in a text console, for example interactive work with the Julia REPL.

Log levels less than min_level are filtered out.

Message formatting can be controlled by setting keyword arguments:

meta_formatter is a function which takes the log event metadata (level, _module, group, id, file, line) and returns a color (as would be passed to printstyled), prefix and suffix for the log message. The default is to prefix with the log level and a suffix containing the module, file and line location.

show_limited limits the printing of large data structures to something which can fit on the screen by setting the :limit IOContext key during formatting.

right_justify is the integer column which log metadata is right justified at. The default is zero (metadata goes on its own line).

MicroLogging.InteractiveLogger — Type.

InteractiveLogger(stream::IO; min_level=Info))

Logger for logging to an interactive terminal, formatting logs in a readable way.

MicroLogging.LogLevel

MicroLogging.MicroLogging

MicroLogging.NullLogger

MicroLogging.SimpleLogger

MicroLogging.configure_logging — Function.

configure_logging(args...; kwargs...)

Call configure_logging with the current logger, and update cached log filtering information.

MicroLogging.core_in_base

MicroLogging.current_logger

MicroLogging.default_logcolor

MicroLogging.default_metafmt

MicroLogging.disable_logging

MicroLogging.display_message

MicroLogging.eval

MicroLogging.formatmsg

MicroLogging.global_logger

MicroLogging.include

MicroLogging.levelstyle

MicroLogging.parse_level

MicroLogging.print_with_decorations

MicroLogging.showvalue

MicroLogging.termlength

MicroLogging.with_logger