Docstrings

BioServices.BioServices

BioServices.EUtils — Module.

Entrez Programming Utilities (or E-Utilities) module.

The EUtils module provides a programming interface to the Entrez databases at NCBI. Nine functions are exported from this module:

einfo: database statistics

esearch: text search

epost: UID upload

esummary: document summary download

efetch: data record download

elink: Entrez link

egquery: global query

espell: spelling suggestion

ecitmatch: batch citation searching in PubMed

See "Entrez Programming Utilities Help" (https://www.ncbi.nlm.nih.gov/books/NBK25501/) for more details. Especially, "E-utilities Quick Start" is a good starting point and "A General Introduction to the E-utilities" will serve useful information about its concepts and functions. The implemented APIs are based on the manual on January 23, 2015.

BioServices.GGGenome — Module.

GGGenome module.

The GGGenome module provides a programming interface to GGGenome, an ultrafast DNA sequence search service, hosted by DBCLS. Two functions are exported from this module:

gggenome: DNA sequence search

gggenome_dblist: List of available databases

See "GGGenome Help" (https://gggenome.dbcls.jp/en/help.html) for more details.

BioServices.UMLS — Module.

BioServices.UMLS

Julia interface to Unified Medical Language REST API

The UMLS provides basic functionality to authenticate and query the UMLS API. The following functions are exported from this module:

get_tgt: Get Ticket-Granting Ticket

search_umls: Search the UMLS

best_match_cui: Get the best matching CUI for a term

get_cui: Get Concept ID

get_semantic_types: Get UMLS semantic type

See "UMLS REST API Home Page" (https://documentation.uts.nlm.nih.gov/rest/home.html) for more details.

BioServices.eval

BioServices.include