Docstrings

ConjugatePriors.ConjugatePriors

ConjugatePriors.Float64WithGamma

ConjugatePriors.Float64WithInverseGamma

ConjugatePriors.MeanAndCovMat

ConjugatePriors.MvNormalWithCov

ConjugatePriors.MvNormalWithCovMat

ConjugatePriors.NormalGamma

ConjugatePriors.NormalInverseChisq — Type.

NormalInverseChisq(μ, σ2, κ, ν)

A Normal-χ^-2 distribution is a conjugate prior for a Normal distribution with unknown mean and variance. It has parameters:

μ: expected mean

σ2 > 0: expected variance

κ ≥ 0: mean confidence

ν ≥ 0: variance confidence

The parameters have a natural interpretation when used as a prior for a Normal distribution with unknown mean and variance: μ and σ2 are the expected mean and variance, while κ and ν are the respective degrees of confidence (expressed in "pseudocounts"). When interpretable parameters are important, this makes it a slightly more convenient parametrization of the conjugate prior.

Equivalent to a NormalInverseGamma distribution with parameters:

m0 = μ

v0 = 1/κ

shape = ν/2

scale = νσ2/2

Based on Murphy "Conjugate Bayesian analysis of the Gaussian distribution".

ConjugatePriors.NormalInverseGamma

ConjugatePriors.NormalInverseWishart

ConjugatePriors.NormalWishart

ConjugatePriors.NormalWithFloat64

ConjugatePriors.complete

ConjugatePriors.eval

ConjugatePriors.fit_map

ConjugatePriors.include

ConjugatePriors.insupport

ConjugatePriors.mu

ConjugatePriors.nu

ConjugatePriors.posterior

ConjugatePriors.posterior_canon

ConjugatePriors.posterior_mode

ConjugatePriors.posterior_rand

ConjugatePriors.posterior_rand!

ConjugatePriors.posterior_randmodel

ConjugatePriors.v0