Docstrings

Robotlib.Ai — Function.

This is a helper method for calibPOE

Robotlib.Calibration

Robotlib.DH

Robotlib.DH2400

Robotlib.DH2twistsLPOE — Function.

Takes the DH-parameters or a set of nominal transformation matrices and outputs the joint twists in local link frames

Robotlib.DH2twistsPOE — Function.

Takes the DH-parameters or a set of nominal transformation matrices and outputs the joint twists in base frame

Robotlib.DH7600

Robotlib.DHYuMi

Robotlib.DHtest

Robotlib.Frames — Module.

Frames is a module to work with coordinate systems. The central type is the Frame with fields R and t for the rotation matrix and translation vector respectively. Other inportnat important types include Point, Points, Plane and Line

All geometric object types have a reference frame associated with them. When a geometric object is created, the reference frame is specified, e.g., T_RB_T = Frame(T_RB_Tm,"RB","T") is a transformation between frame RB and frame T. When transofrmations are made, the reference frames are updated automatically, e.g., T_RB_SEAM = T_RB_T*T_T_SEAM

Usage:

using Robotlib.Frames
path = Pkg.dir("Robotlib","src","applications","frames/")
add_frame_name!("SEAM","Weld seam frame")
add_frame_name!("TAB","Table frame")

T_RB_Tm    = MAT.matread(path*"T_RB_T.mat")["T_RB_T"]
T_TF_TCPm  = MAT.matread(path*"T_TF_TCP.mat")["T_TF_TCP"]
T_T_TABm   = MAT.matread(path*"T_T_Table.mat")["T_T_Table"]

T_RB_T     = Frame(T_RB_Tm,"RB","T")
T_S_D      = Frame(T_TF_TCPm,"S","D")
T_T_TAB    = Frame(T_T_TABm,"T","TAB")

cloud_seam = readcloud(path*"CloudSeam_edge.txt")
plane_seam = readplane(path*"PlaneSeam_edge.txt")
cloud_seam_projected = project(plane_seam,cloud_seam)
line_seam  = fitline(cloud_seam_projected)

T_T_SEAM      = framefromfeatures(("z+",line_seam),("y-",plane_seam),cloud_seam_projected[1],"SEAM")
T_RB_SEAM     = T_RB_T*T_T_SEAM
T_RB_TAB      = T_RB_T*T_T_TAB
T_TAB_SEAM    = inv(T_T_TAB)*T_T_SEAM

cloud_seam_RB = T_RB_T*cloud_seam
cloud_seam_projected_RB = T_RB_T*cloud_seam_projected
plane_seam_RB = T_RB_T*plane_seam
line_seam_RB  = T_RB_T*line_seam

using Plots
default(markersize=1)
plot(Frame(I4,"RB","U"),200, label=true)
plot!(cloud_seam_RB)
plot!(cloud_seam_projected_RB)
plot!(line_seam_RB,500,label="Line seam")
plot!(plane_seam_RB,200,label="Plane seam")
plot!(T_RB_SEAM,200, label=true)
plot!(T_RB_TAB,200, label=true)

xlabel!("x")
ylabel!("y")
zlabel!("z")

MAT.matwrite(path*"T_TAB_SEAM.mat",["T_TAB_SEAM" => T_TAB_SEAM.T])
MAT.matwrite(path*"T_T_SEAM.mat",["T_T_SEAM" => T_T_SEAM.T])
MAT.matwrite(path*"T_RB_TAB.mat",["T_RB_TAB" => T_RB_TAB.T])
println("Wrote T_TAB_SEAM, T_T_SEAM, T_RB_TAB to files in $path")

Robotlib.I3

Robotlib.I4

Robotlib.Plot3Smart

Robotlib.Quaternion

Robotlib.R2rpy — Function.

R2rpy(R; conv="xyz", deg = false)

If conv is not xyz, it will be zyx

returns a vector ∈ R3 or a matrix ∈ R3×N depending on the dimension of the input

Robotlib.Rangle — Function.

Rangle(R1,R2 = I3,deg = false) calculates the angle between two rotation matrices

Robotlib.Robotlib — Module.

This is a library of functions to help out in the Robotlab at LTH The module imports the following files

include("utils.jl")

include("DH.jl")

include("kinematics.jl")

include("robotplot.jl")

include("Frames.jl")

include("read_log.jl")

include("csv2mat.jl")

include("posDepFric.jl")

The module includes a submodule, Frames, which is aimed at replacing the Nikon K600 software. It supports creation of frames, simple projections, fitting of planes, lines etc. and has a number of plotting options. It must be separately imported with using Robotlib.Frames

The module includes a submodule, Calibration, which includes a number of calibration routines. It must be separately imported with using Robotlib.Calibration

Usage: fkine, ikine, jacobian = getkinematicfunctions("yumi") data = orcalog2mat(pathopen, pathsave) q = getData("robot_0.*posRawAbs", data, 1, removeNaN = false)

For YuMi, joint angles q must be converted to logical order using e.g. abb2logical!(q) You must also consider the base transform of YuMi

Robotlib.Rt2T

Robotlib.T2R

Robotlib.T2t

Robotlib.TrajPlot

Robotlib.TrajPlot3

Robotlib.abb2logical

Robotlib.abb2logical!

Robotlib.ad — Function.

Calculates the adjoint of a transformation matrix

Robotlib.adi — Function.

Calculates the adjoint of the inverse of a tranformation matrix T, which is also the inverse of the adjoint of T

Robotlib.basisParametersNd

Robotlib.centralDiff

Robotlib.centraldiff

Robotlib.conformize

Robotlib.csv2mat

Robotlib.dh2Tn

Robotlib.dh2Tn! — Function.

Computes a set of local transformation matrices given the DH-parameters of a robot. Can be sent an optional joint-angle vector and a tool

Robotlib.dir

Robotlib.eval

Robotlib.expξ — Function.

expξ(xi,q=1.0) calculates the exponential map of a twist with twistcoordinates xi and joint angle q If no angle is given, q=1 is assumed

Robotlib.expξ!

Robotlib.expξ2

Robotlib.expξ2! — Function.

Optimized routine to modify T in place

Robotlib.expω

Robotlib.fkineLPOE — Function.

fkineLPOE(Tn0,xi,q) Forward kinematics using LPOE

Robotlib.fkinePOE — Function.

fkinePOE(xi0,q) Forward kinematics using POE

Robotlib.frictionRBFN — Function.

A = frictionRBFN(q, q̇, centers; sigma=zeros(2), normalized=true)

Returns the regressor matrix for position dependent friction estimation

Robotlib.getCenters

Robotlib.getData

Robotlib.get_kinematic_functions — Function.

Run this function with a string representing the robot you want the kinematic functions for, e.g. get_kinematic_functions("YuMi") Currently supports YuMi and ABB IRB7600 returns fkine(q), ikine(T,q0), jacobian(q) TODO: implement YuMileft, Yumiright Tbase*ikinePOE

Robotlib.ikinePOE — Function.

function ikinePOE(xi,T,q0; maxiter=100, λ = 1e0, tol = 1e-12, verbose = false, adaptive = true) Iterative inverse kinematics λsets the start value of the regularizing parameter (kind of like inverse step size). Bigger value means slower but more robust convergence, if adaptive is set to true, λ is adapted to speed up convergence.

Robotlib.include

Robotlib.isrot

Robotlib.isse3

Robotlib.jacobian — Function.

Jn, J0, T, Ti, Tn = jacobian(qin, DH::DH, tool=I4) Calculates the jacobian in base (J0) and tool frame (Jn) as well as the forward kinematics (T), given the joint angles and the DH-parameters

Robotlib.jacobianPOE — Function.

jacobianPOE(q, xi) Returns The jacobian in 1) the base frame, 2) the tool frame. It can most likely be rewritten to be faster.

Robotlib.jacobianPOEb

Robotlib.jacobianPOEikine — Function.

jacobianPOE(q, xi, T) Returns The jacobian in 1) the base frame, 2) the tool frame and returns the FK. It must be handed a precomputed end transform Tend = expξ2(xi[:,end-1],1)*expξ2(xi[:,end],1)

Robotlib.logR — Function.

Calculates the matrix logarithm of a rotation matrix ∈ SO(3)

Robotlib.logT — Function.

Calculates the matrix logarithm of a transformation matrix ∈ SE(3) Does not seem to be very reliable for very small rotations, use logm instead, which is a bit slower.

Robotlib.logical2abb

Robotlib.logical2abb!

Robotlib.mnorm_pdf

Robotlib.orcalog2mat

Robotlib.plot3smart — Function.

plot3smart(x,args...) Makes a 3d plot of a matrix

Robotlib.plot3smart!

Robotlib.prodad

Robotlib.readmat

Robotlib.rotx

Robotlib.roty

Robotlib.rotz

Robotlib.rpy2R — Function.

rpy2R(r,p,y,conv="zyx") For rpy from ABB robot, use zyx For rpy from Nikon, use xyz

Robotlib.skew

Robotlib.skew!

Robotlib.skew4

Robotlib.skewcoords

Robotlib.smartDiff

Robotlib.testJacobian

Robotlib.testikine

Robotlib.toOrthoNormal

Robotlib.toOrthoNormal! — Function.

Takes a matrix R ∈ SO(3) or T ∈ SE(3) and makes the rotational part orthonormal

Robotlib.traj2quat

Robotlib.trajplot — Function.

trajplot(T,args...) Plots a trajectory of T-matrices in a single plot

Robotlib.trajplot!

Robotlib.trajplot3 — Function.

plot_traj3(T, ls=".", plotFrame = 0) Plots a trajectory of T-matrices in a 3D plot, with otional frames drawn of length plotFrame

Robotlib.trajplot3!

Robotlib.trinv — Function.

Inverts a transformation matrix ∈ SE(3)

Robotlib.twistcoords

Robotlib.xii — Function.

This is the other helper method for calibPOE

Robotlib.xyθ — Function.

Calculate the (signed) angle around z axis for two rotation (transformation) matrices.