Docstrings

PicoSAT.PicoPtr

PicoSAT.PicoSAT

PicoSAT.PicoSolIterator

PicoSAT.SATISFIABLE

PicoSAT.UNKNOWN

PicoSAT.UNSATISFIABLE

PicoSAT.add_clause

PicoSAT.add_clauses

PicoSAT.blocksol

PicoSAT.config

PicoSAT.copyright

PicoSAT.eval

PicoSAT.get_solution

PicoSAT.include

PicoSAT.itersolve — Function.

itersolve(clauses; vars::Integer=-1, verbose::Integer=0, proplimit::Integer=0)

vars - the number of variables

verbose - prints solver logs to STDOUT when verbose > 0 with increasing detail.

proplimit - helps to bound the execution time. The number of propagations and the solution time are roughly linearly related. A value of 0 (default) allows for an unbounded number of propagations.

Returns an iterable object over all solutions. When a user-defined propagation limit is specified, the iterator may not produce all feasible solutions.

julia> import PicoSAT
julia> cnf = Any[[1, -5, 4], [-1, 5, 3, 4], [-3, -4]];
julia> PicoSAT.itersolve(cnf)
julia> for sol in PicoSAT.itersolve(cnf)
           println(sol)
       end
[1,-2,-3,-4,5]
[1,-2,-3,4,-5]
[1,-2,-3,4,5]
[1,-2,3,-4,-5]
...

PicoSAT.libpicosat

PicoSAT.next_solution

PicoSAT.picosat_add — Function.

Add a literal of the next clause. A zero terminates the clause. The solver is incremental. Adding a new literal will reset the previous assignment. The return value is the original clause index to which this literal respectively the trailing zero belong starting at 0.

PicoSAT.picosat_adjust — Function.

If you know a good estimate on how many variables you are going to use then calling this function before adding literals will result in less resizing of the variable table. But this is just a minor optimization. Beside exactly allocating enough variables it has the same effect as calling 'picosatincmax_var'.

PicoSAT.picosat_deref — Function.

After 'picosatsat' was called and returned 'PICOSATSATISFIABLE', then the satisfying assignment can be obtained by 'dereferencing' literals. The value of the literal is return as '1' for 'true', '-1' for 'false' and '0' for an unknown value.

PicoSAT.picosat_init

PicoSAT.picosat_measure_all_calls — Function.

Measure all time spent in all calls in the solver. By default only the time spent in 'picosatsat' is measured. Enabling this function might for instance triple the time needed to add large CNFs, since every call to 'picosatadd' will trigger a call to 'getrusage'.

PicoSAT.picosat_reset

PicoSAT.picosat_sat — Function.

Call the main SAT solver. A negative decision limit sets no limit on the number of decisions.

PicoSAT.picosat_set_plain — Function.

Disable (set_plain == true) / Enable all prerocessing.

PicoSAT.picosat_set_prefix — Function.

Set the prefix used for printing verbose messages and statistics. (Default is "c ")

PicoSAT.picosat_set_propagation_limit — Function.

As alternative to a decision limit you can use the number of propagations as limit. This is more linearly related to execution time. This has to be called after 'picosatinit' and before 'picosatsat'.

PicoSAT.picosat_set_verbosity — Function.

Set verbosity level A verbosity level >= 1 prints more and more detailed progress reports to the output file. Verbose messages are prefixed with the string set by 'picosatsetprefix'

PicoSAT.picosat_setup

PicoSAT.picosat_variables — Function.

p cnf <m> n

PicoSAT.satisfiable

PicoSAT.solve — Function.

solve(clauses; vars::Integer=-1, verbose::Integer=0, proplimit::Integer=0)

vars - the number of variables

verbose - prints solver logs to STDOUT when verbose > 0 with increasing detail.

proplimit - helps to bound the execution time. The number of propagations and the solution time are roughly linearly related. A value of 0 (default) allows for an unbounded number of propagations.

Returns a solution if the problem is satisfiable. Satisfiable solutions are represented as a vector of signed integers. If the problem is not satisfiable the method returns an :unsatisfiable symbol. If a solution cannot be found within the defined propagation limit, an :unknown symbol is returned.

julia> import PicoSAT
julia> cnf = Any[[1, -5, 4], [-1, 5, 3, 4], [-3, -4]];
julia> PicoSAT.solve(cnf)
5-element Array{Int64,1}:
   1
  -2
  -3
  -4
   5

PicoSAT.version