Docstrings

CausalityToolsBase.ApproximateIntersection — Type.

ApproximateIntersection

A type indicating that simplex intersections should be computed exactly.

CausalityToolsBase.BinningScheme — Type.

BinningScheme

The supertype of all binning schemes in the CausalityTools ecosystem.

CausalityToolsBase.CausalityToolsBase

CausalityToolsBase.CustomReconstruction

CausalityToolsBase.ExactIntersection — Type.

ExactIntersection

A type indicating that simplex intersections should be computed exactly.

CausalityToolsBase.Lags — Type.

Lags

Wrapper type for lags used when performing custom state space reconstructions. Used in combination with Positions to specify how a CustomReconstruction should be constructed.

Examples

Lags(2, 0, -3, 1) indicates a 4-dimensional state space reconstruction where the first variable has a positive lag of 2, the second variable is not lagged, the third variable has a lag of -3, and the fourth variable has a positive lag of 1.

Lags(0, 0) indicates a 2-dimensional state space reconstruction where both variables are not lagged.

CausalityToolsBase.Positions — Type.

Positions

Wrapper type for the positions the different dynamical variables appear in when constructing a custom state space reconstruction. Used in combination with Lags to specify how a CustomReconstruction should be constructed. Each of the positions must refer to a dynamical variable (column) actually present in the dataset.

Examples

Positions(1, 2, 1, 5) indicates a 4-dimensional state space reconstruction where

the 1st coordinate axis of the reconstruction should be formed from the

first variable/column of the input data.

the 2nd coordinate axis of the reconstruction should be formed from the

2nd variable/column of the input data.

the 3rd coordinate axis of the reconstruction should be formed from the

1st variable/column of the input data.

the 4th coordinate axis of the reconstruction should be formed from the

5th variable/column of the input data.

Positions(-1, 2) indicates a 2-dimensional reconstruction, but will not work, because each position must refer to the index of a dynamical variable (column) of a dataset (indexed from 1 and up).

CausalityToolsBase.ReconstructionParameters

CausalityToolsBase.RectangularBinning — Type.

RectangularBinningScheme

Instructions for creating a rectangular box partition.

Fields

ϵ::Union{Int, Float64, Vector{Int}, Vector{Float64}}: The instructions for deciding the edge lengths of the rectangular boxes. The following ϵ are valid:

  1. `ϵ::Int` divides each axis into `ϵ` intervals of the same size.

  2. `ϵ::Float` divides each axis into intervals of size `ϵ`.

  3. `ϵ::Vector{Int}` divides the i-th axis into `ϵᵢ` intervals of the same size.
  
  4. `ϵ::Vector{Float64}` divides the i-th axis into intervals of size `ϵᵢ`.

CausalityToolsBase.RectangularBinningScheme — Type.

RectangularBinningScheme

The supertype of all rectangular binning schemes in the CausalityTools ecosystem.

CausalityToolsBase.RefinedTriangulationBinningMaxRadius — Type.

RefinedTriangulationBinningMaxRadius

A binning scheme for a triangulated simplex partition where some simplices have been refined (subdivided by a shape-preserving simplex subdivision algorithm).

The maximum radius bound is applied by first doing an initial triangulation, the splitting simplices whose radius is large until all simplices have radii less than the resulting radius bound.

Fields

max_radius_frac::Float64: The maximum radius expressed as a fraction of the

radius of the largest simplex of the initial triangulation.

CausalityToolsBase.RefinedTriangulationBinningSplitFactor — Type.

RefinedTriangulationBinningSplitFactor

A binning scheme for a triangulated simplex partition where some simplices have been refined (subdivided by a shape-preserving simplex subdivision algorithm).

The split factor bound controls how many times each simplex of the initial triangulation is to be split.

Fields

simplex_split_factor::Int: The number of times each simplex is split.

CausalityToolsBase.RefinedTriangulationBinningSplitQuantile — Type.

RefinedTriangulationBinningSplitQuantile

A binning scheme for a triangulated simplex partition where some simplices have been refined (subdivided by a shape-preserving simplex subdivision algorithm).

The split fraction bound controls how many times each simplex of the initial triangulation is to be split.

Fields

split_quantile::Float64: All simplices with radius larger than the split_quantile-th quantile of the radii of the simplices in initial triangulation are split with a splitting factor of simplex_split_factor.

simplex_split_factor::Int: The number of times each simplex is split.

CausalityToolsBase.SimplexIntersectionType — Type.

SimplexIntersectionType

An abstract type for different types of simplex intersections.

CausalityToolsBase.TriangulationBinning — Type.

TriangulationBinningScheme

A binning scheme for a triangulated simplex partition.

CausalityToolsBase.TriangulationBinningScheme — Type.

TriangulationBinningScheme

The supertype of all triangulation binning schemes in the CausalityTools ecosystem.

CausalityToolsBase.customembed — Function.

customembed(pts, positions::Positions, lags::Lags)

Do custom state space reconstructions with customembed(pts, positions::Positions, lags::Lags). This function acts almost as DynamicalSystems.reconstruct, but allows for more flexibility in the ordering of dynamical variables and allows for negative lags. The positions variable indicates which dynamical variables are mapped to which variables in the final reconstruction, while lags indicates the lags for each of the embedding variables.

Example: customembed([rand(3) for i = 1:50], Positions(1, 2, 1, 3), Lags(0, 0, 1, -2) gives a 4-dimensional embedding with state vectors (x1(t), x2(t), x1(t + 1), x3(t - 2)).

Note: customembed expects an array of state vectors, i.e. pts[k] must refer to the kth point of the dataset, not the kth dynamical variable/column.*. To embed a vector of time series, load DynamicalSystems and wrap the time series in a Dataset first, e.g. if x = rand(100); y = rand(100) are two time series, then customembed(Dataset(x, y), Positions(1, 2, 2), Lags(0, 0, 1) will create the embedding with state vectors (x(t), y(t), y(t + 1)).

Pre-embedded points may be wrapped in a CustomReconstruction instance by simply calling customembed(preembedded_pts) without any position/lag instructions.

CausalityToolsBase.encode — Function.

encode(point, reference_point, edgelengths)

Encode a point into its integer bin labels relative to some reference_point (always counting from lowest to highest magnitudes), given a set of box edgelengths (one for each axis). The first bin on the positive side of the reference point is indexed with 0, and the first bin on the negative side of the reference point is indexed with -1.

CausalityToolsBase.eval

CausalityToolsBase.fill_embedding_pts!

CausalityToolsBase.get_edgelengths — Function.

get_edgelengths(pts, ϵ) -> Vector{Float}

Return the box edge length along each axis resulting from discretizing pts on a rectangular grid specified by the binning scheme ϵ.

CausalityToolsBase.get_maxima

CausalityToolsBase.get_minima — Function.

get_maxima(pts) -> Vector{Float}

Return the maxima along each axis of the dataset pts.

CausalityToolsBase.get_minima_and_edgelengths — Function.

get_minima_and_edgelengths(points, ϵ) -> (Vector{Float}, Vector{Float})

Find the minima along each axis of the embedding, and computes appropriate stepsizes given ϵ, which provide instructions on how to grid the space. Assumes each point is a column vector.

Specifically, the binning procedure is controlled by the type of ϵ:

ϵ::Int divides each axis into ϵ intervals of the same size.

ϵ::Float divides each axis into intervals of size ϵ.

ϵ::Vector{Int} divides the i-th axis into ϵᵢ intervals of the same size.

ϵ::Vector{Float64} divides the i-th axis into intervals of size ϵᵢ.

CausalityToolsBase.get_minmaxes — Function.

get_minmaxes(pts) -> Tuple{Vector{Float}, Vector{Float}}

Return a vector of tuples containing axis-wise (minimum, maximum) values.

CausalityToolsBase.include

CausalityToolsBase.joint_visits — Function.

joint_visits(points, ϵ)

Determine which bins are visited by points given the rectangular binning scheme ϵ. Bins are referenced relative to the axis minimum.

CausalityToolsBase.marginal_visits — Function.

marginal_visits(points, ϵ, dims)

Determine which bins are visited by points given the rectangular binning scheme ϵ, only along the desired dimensions dims. Bins are referenced relative to the axis minimum.

marginal_visits(joint_visits, dims)

Given a set of precomputed joint visited bins, return the marginal along dimensions dims.

ChaosTools.non0hist — Function.

non0hist(bin_visits)

Return the unordered histogram (vistitation frequency) over the array of bin_visits, which is a vector containing bin encodings.

This method extends ChaosTools.non0hist.

non0hist(points, ϵ, dims)

Determine which bins are visited by points given the rectangular binning scheme ϵ, considering only the marginal along dimensions dims. Bins are referenced relative to the axis minimum.

Returns the unordered histogram (vistitation frequency) over the array of bin visits.

This method extends ChaosTools.non0hist.

CausalityToolsBase.optimal_delay — Function.

optimal_delay(v; method = "mi_min")

Estimate the optimal embedding delay for v using the provided method. See the documentation for DelayEmbeddings.estimate_delay for a list of available methods.

CausalityToolsBase.optimal_dimension — Function.

optimal_dimension(v, τ; dims = 1:8; method = "fnn"; kwargs...)

Estimate the optimal embedding dimension for v.

Arguments

v: The data series for which to estimate the embedding dimension.

τ: The embedding lag.

dims: Dimensions to probe for the optimal dimension.

Keyword arguments

method: Either "fnn" (Kennel's false nearest neighbors method), "afnn" (Cao's average false nearest neighbors method) or "f1nn" (Krakovská's false first nearest neighbors method). See the source code for DelayEmbeddings.estimate_dimension for more details.

rtol: Tolerance rtol in Kennel's algorithms. See DelayEmbeddings.fnn source code for more details.

atol: Tolerance rtol in Kennel's algorithms. See DelayEmbeddings.fnn source code for more details.

optimal_dimension(v; dims = 1:8,
    method_dimension = "fnn", method_delay = "first_min")

Estimate the optimal embedding dimension for v by first estimating the optimal lag, then using that lag to estimate the dimension.

Arguments

v: The data series for which to estimate the embedding dimension.

dims: The dimensions to try

CausalityToolsBase.verify_valid_positions!