Docstrings

Permutations.AbstractPermutation

Permutations.CoxeterDecomposition — Type.

CoxeterDecomposition(p) expresses the Permutation p as a composition of transpositions of the form (k,k+1).

Permutations.CoxeterDecomposition!

Permutations.CoxeterGenerator

Permutations.Permutation — Type.

Permutation(list) creates a new Permutation. Here list must be a rearrangement of 1:n.

Permutation(n) creates the identity Permutation of 1:n.

Permutation(n,k) creates the k'th Permutation of 1:n.

Permutations.Permutations

Permutations.RandomPermutation — Function.

RandomPermutation(n) creates a random permutation of 1:n, each with probability 1/factorial(n).

Permutations._coxeter_reduce!

Permutations._coxeterdecomposition!

Permutations.array

Permutations.array2string

Permutations.cycle_string — Function.

cycle_string(p) creates a nice, prinatble string representation from the cycle structure of the permutation p.

Permutations.cycles — Function.

cycles(p) returns a list of the cycles in Permutation p.

Permutations.dict — Function.

dict(p) converts a permutation into a dictionary. If d is the result then d[k] equals p(k) for all k in the domain of the permutation.

Permutations.eval

Permutations.fixed_points — Function.

fixed_points(p) returns the list of values k for which p(k)==k.

Permutations.getindex

Permutations.hash

Permutations.include

Base.inv — Function.

inv(p) gives the inverse of Permutation p. This may also be computed with p'.

Base.length — Function.

length(p) is the number of elements in the Permutation p.

Permutations.longest_decreasing — Function.

Let p be a Permutation. Thinking of this as a list of values, longest_decreasing(p) returns a longest subsequence of values that are in descending order

Permutations.longest_increasing — Function.

Let p be a Permutation. Thinking of this as a list of values, longest_increasing(p) returns a longest subsequence of values that are in ascending order

Permutations.longest_monotone

Permutations.matrix

Permutations.order — Function.

order(p) is the smallest positive integer n such that p^n is the identity Permutation.

Base.reverse — Function.

Let p be a Permutation. Thinking of p as a list of values, reverse(p) creates a Permutation with those values in reverse sequence.

Base.sign — Function.

sign(p) is +1 is p is an even Permtuation and -1 if p is odd.

Permutations.sort_by_size — Function.

sort_by_size(list) returns a sorted copy of list in which the elements are sorted by the length function (from smallest to largest).

Permutations.two_row — Function.

two_row(p) creates a two-row representation of the Permutation p in which the first row is 1:n and the second row are the values p(1), p(2), ..., p(n).

Permutations.weave