Docstrings

ObjectFile.@constants — Macro.

@constants

Macro to create intelligent enum arrays in Julia. Defines not only variables mapping names to values within Julia, but also a dictionary mapping those values back to a string representation of the variable names itself.

ObjectFile.@derefmethod — Macro.

@derefmethod

Macro to create a method that works on a reference type by generating a wrapper call to deref(x) where x is the first argument in the call. Example:

@derefmethod foo(x::SectionRef)

Will generate the following code:

foo(x::SectionRef, args...) = foo(deref(x), args...)

ObjectFile.@mustimplement — Macro.

@mustimplement

Macro to create fallthrough implementations of basic functions such as readheader(oh::ObjectHandle); these fallthrough implementations are meant to be overridden by methods in packages such as ELF.jl or MachO.jl.

ObjectFile.COFF

ObjectFile.COFFDataDirectories

ObjectFile.COFFDynamicLink

ObjectFile.COFFDynamicLinks

ObjectFile.COFF.COFFHandle — Type.

COFFHandle

An ObjectHandle subclass for COFF files, this is the primary object by which client applications will interact with COFF files.

ObjectFile.COFFHeader

ObjectFile.COFFImageDataDirectory

ObjectFile.COFFOptionalHeader32

ObjectFile.COFFOptionalHeader64

ObjectFile.COFFOptionalHeaderStandard

ObjectFile.COFFOptionalHeaderWindows32

ObjectFile.COFFOptionalHeaderWindows64

ObjectFile.COFF.COFFRPath — Type.

COFFRPath

COFF RPath object; note that while COFF files do not have an RPath within them, they do seach the same directory as the loading binary (e.g. the $ORIGIN). We use COFFRPath to effect this, although strictly speaking there is no such thing as a "COFF RPath".

ObjectFile.COFF.COFFStrTab — Type.

COFFStrTab

COFF StrTab type, containing the metadata necessary to perform string table lookups, via the strtab_lookup() method.

ObjectFile.COFF.COFFSymbolRef — Type.

COFFSymbolRef

Contains a reference to an COFFSymtabEntry, as well as an COFFSymbols, etc...

ObjectFile.COFF.COFFSymbols — Type.

COFFSymbols

COFF symbol table, contains the list of symbols defined within the object file.

Note that because COFF Symbols are variable-length, we store a table of offsets at which the (non-auxilliary) symbols can be found.

ObjectFile.COFFSymtabEntry

ObjectFile.DynamicLink — Type.

DynamicLink

This type encapsulates the linkage of one object file to another. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

DynamicLink()

Accessors:

DynamicLinks()

handle()

path()

ObjectFile.DynamicLinks — Type.

DynamicLinks

This type encapsulates the list of dynamic links within an object, holding a collection of DynamicLink objects. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

DynamicLinks()

Iteration

getindex()

lastindex()

iterate()

eltype()

Misc.

handle()

ObjectFile.ELF

ObjectFile.ELF.ELFDynEntries — Function.

ELFDynEntries(oh::ELFHandle)

Read all ELFDynEntry objects from an ELF object, returning them as an array.

ELFDynEntries(oh::ELFHandle, kinds::Vector)

Read all ELFDynEntry objects from an ELF object, returning them as an array if they are one of the kinds passed in, such as DT_NEEDED.

ObjectFile.ELF.ELFDynEntry — Type.

ELFDynEntry

ELF Dynamic table entry type. This is comprised by a d_tag member and a d_val member; the d_tag tells what kind of command this is to the dynamic linker (e.g. DT_NEEDED denotes a shared library that must be loaded for this ELF object to link properly), whereas d_val is a pointer to another data structure that contains more information. (E.g. for a DT_NEEDED entry, d_val would represent an offset within the dynamic string table)

ObjectFile.ELFDynEntry32

ObjectFile.ELFDynEntry64

ObjectFile.ELFDynEntryRef

ObjectFile.ELFDynamicLink

ObjectFile.ELF.ELFDynamicLinks — Type.

ELFDynamicLinks

ELF dynamic linkage table, contains the list dynamic loader commands, see the ELFDynEntry type for an opaque and unhelpful detailing of these commands.

ObjectFile.ELF.ELFHandle — Type.

ELFHandle

An ObjectHandle subclass for ELF files, this is the primary object by which client applications will interact with ELF files.

ObjectFile.ELF.ELFHeader — Type.

ELFHeader

ELF Header type, containing metadata about the ELF object itself, such as its type, offsets for the Program and Section headers, the number of other header entries, etc...

ObjectFile.ELFHeader32

ObjectFile.ELFHeader64

ObjectFile.ELF.ELFInternal — Type.

ELFInternal

Internal datastructure used within the ELF file format to convey fundamental information such as the endianness of the file, whether it's a 32-bit or 64-bit ELF file, etc...

ObjectFile.ELF.ELFRPath — Type.

ELFRPath

Stores the RPath entries from an ELF object

ObjectFile.ELF.ELFSection — Type.

ELFSection

ELF Section type, containing information about a Section within the ELF object, such as the Section's name, its size, etc...

ObjectFile.ELFSection32

ObjectFile.ELFSection64

ObjectFile.ELF.ELFSectionRef — Type.

ELFSectionRef

ELF SectionRef type, containing an ELFSection and important metadata.

ObjectFile.ELF.ELFSections — Type.

ELFSections

ELF Section header table type, containing information about the number of sections within the ELF object, the location of the section headers, etc...

ObjectFile.ELF.ELFSegment — Type.

ELFSegment

ELF Segment type, also known as a program header, containing information about the virtual memory layout of a chunk of the program.

ObjectFile.ELFSegment32

ObjectFile.ELFSegment64

ObjectFile.ELF.ELFSegments — Type.

ELFSegments

ELF segment header table type, containing information about the number of segments within the ELF object, the location of the segment headers, etc...

ObjectFile.ELF.ELFStrTab — Type.

ELFStrTab

ELF string table type, containing information about a Section that contains a string table. Note that an ELF file may contain multiple string tables, use the different constructors to reference the different string tables.

ObjectFile.ELF.ELFSymbolRef — Type.

ELFSymbolRef

Contains a reference to an ELFSymtabEntry, as well as an ELFSymbols, etc...

ObjectFile.ELF.ELFSymbols — Type.

ELFSymbols

ELF symbol table, contains the list of symbols defined within the object file.

ObjectFile.ELF.ELFSymtabEntry — Type.

ELFSymtabEntry

ELF Symtab Entry type, contains the data relevant to a symbol defined within this ELF file, garnered from the symbol table (symtab)

ObjectFile.ELFSymtabEntry32

ObjectFile.ELFSymtabEntry64

ObjectFile.FatMachOHandle

ObjectFile.MachFatArch

ObjectFile.MachFatHeader

ObjectFile.MachHeader

ObjectFile.MachHeader32

ObjectFile.MachHeader64

ObjectFile.MachO

ObjectFile.MachO.MachODynamicLink — Type.

MachODynamicLink

MachO type representing a dynamic link between a MachO file and one of its dynamic dependencies. Although Mach-O encodes more than just the path of the dependency, in order to get at it you will need to dig into the LoadCmd that describes it.

ObjectFile.MachODynamicLinks

ObjectFile.MachOHandle

ObjectFile.MachO.MachOIdDylibCmd — Type.

MachOIdDylibCmd

MachO dylibs have an "identity", analogous to the SONAME of an ELF shared library. This Load Command tells the linker about the identity of this particular shared library. The API of this object is identical to that of the MachOLoadDylibCmd object, as they are intrinsically identical, it is the semantic meaning behind the data that changes, and so this object is all but a typealias for MachOLoadDylibCmd.

ObjectFile.MachO.MachOLoadCmd — Type.

MachOLoadCmd

Abstraction over all Load Commands that can exist within a Mach-O object file, containing subclasses such as MachOSegmentCmd or MachODySymtabCmd. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

MachOLoadCmd()

ObjectFile.MachO.MachOLoadCmdHeader — Type.

MachOLoadCmdHeader

All MachO Load Commands have a common header, containing information about what kind of Load Command it is, and how large the load command is.

source

ObjectFile.MachO.MachOLoadCmdRef — Type.

MachOLoadCmdRef

Datatype containing a Mach-O Load Command, its header, and a reference back to the MachOLoadCmds object it was garnered from

ObjectFile.MachO.MachOLoadCmds — Type.

MachOLoadCmds

Allows iteration over the LoadCmds within a Mach-O file.

Creation

MachOLoadCmds()

Iteration

getindex()

lastindex()

iterate()

length()

eltype()

Access

find

Convenience constructors

Segments()

ObjectFile.MachO.MachOLoadDylibCmd — Type.

MachOLoadDylibCmd

The Load Command that gives information about a dylib that must be loaded for this Mach-O file to link properly. The API for this laod command is given as follows:

Creation:

MachOLoadCmd()

Accessors:

dylib_name()

dylib_timestamp()

dylib_version()

dylib_compatibility()

ObjectFile.MachORPath

ObjectFile.MachO.MachOSection — Type.

MachOSection

Mach-O Section type, containing information about the section name, segment name, size, address, etc... of the Section.

ObjectFile.MachO.MachOSectionRef — Type.

MachOSectionRef

Mach-O SectionRef type

ObjectFile.MachO.MachOSections — Type.

MachOSections

Mach-O Section object collection type. Mach-O sections are split out over multiple segments, one per load command. As an example, most executables have at least two segments, __DATA and __TEXT, each of which have multiple sections embedded within them. A MachOSections object is created from multiple segments (in this case, realized as multiple load commands) which contain the necessary sections, and these sections can then be accessed as desired.

ObjectFile.MachO.MachOSegment — Type.

MachOSegment

MachO Segment type which is returned from a MachOSegmentRef when it is dereferenced. Note that this type exists only for the purposes of type conformity; all the real work is done within MachOSegmentCmd; but since that already inherits from MachOLoadCmd, it cannot also inherit from Segment. Thus, this wrapper type was born to bridge the type hierarchy.

ObjectFile.MachOSegment32Cmd

ObjectFile.MachOSegment64Cmd

ObjectFile.MachO.MachOSegmentCmd — Type.

MachOSegmentCmd

Mach-O Segment load command type, containing information about the virtual memory layout of a chunk of the program. This type is very MachO-specific, so it is not comparable to other object file formats directly. However, note that this data structure is the gateway through which the file sections are accessed, and while there is a convenience Sections(::MachOHandle) method that will abstract all that away for you, you can also directly call Sections(::MachOLoadCmdRef{MachOSegmentCmd}) to get the sections belonging to the given segment. Note that the Sections call works only upon a MachOLoadCmdRef{MachOSegmentCmd}, it will not work on the MachOSegmentCmd itself directly.

Creation:

MachOSegmentCmd()

Format-specific properties:

segment_name()

segment_offset()

segmentfilesize()

segmentmemorysize()

segmentnumsections()

ObjectFile.MachO.MachOSegmentRef — Type.

MachOSegmentRef

MachO SegmentRef type which is returned from a MachOSegments when it is indexed into. Note that this type exists only for the purposes of type conformity; all the real work is done within MachOSegmentCmd; but since that already inherits from MachOLoadCmd, and we needed a MachOSegment type that would inherit from Segment, we figured might as well make a MachOSegmentRef as well to keep things nice and symmetric.

ObjectFile.MachO.MachOSegments — Type.

MachOSegments

Segments type that contains all load commands within a Mach-O file that contain segment commands.

ObjectFile.MachOSymbolRef

ObjectFile.MachO.MachOSymbols — Type.

MachOSymbols

MachO container type for SymtabEntry objects.

ObjectFile.MachO.MachOSymtabEntry — Type.

MachOSymtabEntry

MachO Symbol Table entry type, internally represented as an nlist type.

ObjectFile.MagicMismatch — Type.

MagicMismatch

This is an error type used to denote that readmeta() was called on a file that does not contain the proper magic at the beginning for the type of object file that was attempting to be loaded.

ObjectFile.ObjTypes — Constant.

ObjTypes::Vector{Type}

ObjTypes contains the registry of file formats that will be used to try and open a object file, (e.g. ELF, MachO, etc...). This vector is initialized at __init__() time, and used within readmeta().

ObjectFile.ObjectFile

ObjectFile.ObjectHandle — Type.

ObjectHandle

The basic type that provides access to object files. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis. Note tha "must implement" is a bit of a misnomer, if an Object file does not have need of a certain piece of this API (e.g. COFF files have no concept of Segments), leaving that chunk of the API undefined will simply cause errors if a user attempts to use methods that use that part of the API (in the example above, an error will be thrown if the user calls Segments(oh) where oh <: COFFHandle).

Creation

readmeta()

IOStream-like operations:

seek()

seekstart()

skip()

startaddr()

iostream()

position()

read()

readuntil()

eof()

unpack()

Format-specific properties

endianness()

is64bit()

isrelocatable()

isexecutable()

islibrary()

isdynamic()

manglesectionname()

manglesymbolname()

handle()

header()

format_string()

Section properties

sectionheaderoffset()

sectionheadersize()

sectionheadertype()

Segment properties

segmentheaderoffset()

segmentheadersize()

segmentheadertype()

Symbol properties

symtabentryoffset()

symtabentrysize()

symtabentrytype()

Misc

path()

show()

find_library()

find_libraries()

ObjectFile.RPath — Type.

RPath

This type encapsulates the search path used by an object file when looking for a shared library. This class enables not only looking at the path, but querying the path for matches for given library names. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

RPath()

Utility

handle()

RPath operations

rpaths()

canonical_rpaths()

find_library()

lastindex()

iterate()

ObjectFile.Section — Type.

Section

An abstraction over the concept of a Section within an object file. Because many operations upon sections require global operations (access to the string table, knowledge of position within the file, etc...) some operations are defined only upon the SectionRef datatype. As a user, the SectionRef type should be the primary method of interacting with sections, as a developer adding new object file formats, some methods must support Sections, others must support only SectionRefs. Note that any method that works on a Section must also work with a SectionRef, see the @derefmethod macro for a convenient helper macro to generate SectionRef -> Section wrapper methods. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

read()

Utility:

deref()

IO-like operations:

contents()

Format-specific properties:

section_name()

section_size()

section_offset()

section_address()

ObjectFile.SectionRef — Type.

SectionRef

Provides a reference to a Section, along with a reference to the ObjectHandle this Section comes from. This should be the primary method by which users interact with sections inside object files. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis. Note that this overlaps heavily with the Section object API, this is by design as many of the methods are simply passthroughs to the underlying Section API calls for ease of use.

Creation:

SectionRef()

Utility

deref()

handle()

Sections()

IO-like operations:

read()

seekstart()

seek()

eof()

Format-specific properties:

section_name()

section_number()

section_type()

section_size()

section_offset()

section_address()

ObjectFile.Sections — Type.

Sections

An abstraction over the concept of a collection of Section types within an object file. One can think of the Sections object containing the table of section headers within the object file, whereas the Section/SectionRef objects contain the actual section data itself. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

Sections()

Iteration

getindex()

lastindex()

length()

iterate()

eltype()

Search

findall()

findfirst()

Misc.

handle()

ObjectFile.Segment — Type.

Segment

An abstraction over the concept of a Segment within an object file. A Segment is a portion of an object file that is given instruction on its layout in virtual memory; this is in contrast to a Section, which delineates different portions of an object file on disk. ELF files have the strictest separation here, with a single executable file containing multiple Segment and Section objects, with Sections being assigned to one or more Segments for virtual memory placement. Mach-O files typically have two Segments, one called __TEXT, one called __DATA. COFF files do not have Segment.

Just like with Section objects, many operations upon segments require global operations (access to the string table, knowledge of position within the file, etc...) which causes some operations to be defined only upon the SegmentRef datatype. As a user, the SegmentRef type should be the primary method of interacting with segments, as a developer adding new object file formats, some methods must support Segments, others must support only SegmentRefs. Note that any method that works on a Segment must also work with a SegmentRef, see the @derefmethod macro for a convenient helper macro to generate SegmentRef -> Section wrapper methods. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

read()

Utility:

deref()

Format-specific properties:

segment_name()

segment_offset()

segmentfilesize()

segmentmemorysize()

segment_address()

ObjectFile.SegmentRef — Type.

SegmentRef

Provides a reference to a Segment, along with a reference to the ObjectHandle this Segment comes from. This should be the primary method by which users interact with segments inside object files. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis. Note that this overlaps heavily with the Segment object API, this is by design as many of the methods are simply passthroughs to the underlying Segment API calls for ease of use.

Creation:

SegmentRef()

Utility

deref()

Segments()

handle()

Format-specific properties:

segment_name()

segment_number()

segment_offset()

segmentfilesize()

segmentmemorysize()

segment_address()

ObjectFile.Segments — Type.

Segments

An abstraction over the concept of a collection of Segment types within an object file. One can think of the Segments object containing the table of segment headers within the object file, whereas the Segment/SegmentRef objects contain the actual segment data itself. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

Segments()

Iteration

getindex()

lastindex()

length()

iterate()

eltype()

Search

findall()

findfirst()

Misc.

handle()

ObjectFile.StrTab — Type.

StrTab

This type encapsulates a string table within an object file, enabling queries against the string table for symbol names, section names, etc... The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

StrTab()

Accessors

handle()

strtab_lookup()

ObjectFile.SymbolRef — Type.

SymbolRef

Provides a reference to a SymtabEntry, along with a reference to the ObjectHandle this SymtabEntry comes from. This should be the primary method by which users interact with symbols inside object files. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis. Note that this overlaps heavily with the SymtabEntry object API, this is by design as many of the methods are simply passthroughs to the underlying SymtabEntry API calls for ease of use.

Creation:

SymbolRef()

Util:

deref()

Symbols()

handle()

Properties:

symbol_number()

symbol_name()

symbol_value()

isundef()

ObjectFile.Symbols — Type.

Symbols

An abstraction over the concept of a collection of symbol (SymtabEntry) types within an object file. One can think of the Symbols object containing the table of symbols within the object file, whereas the SymtabEntry/SymbolRef objects contain the actual symbol data itself. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

Symbols()

Iteration

getindex()

lastindex()

length()

iterate()

eltype()

Misc.

handle()

ObjectFile.SymtabEntry — Type.

SymtabEntry

An abstraction over the concept of a symbol within an object file. This type does not use the Symbol name as this would conflict with the builtin Julia Symbol type, so the name SymtabEntry is used instead. As a user, the SymbolRef type should be the primary method of interacting with symbols, as a developer adding new object file formats, some methods must support SymtabEntrys, others must support only SymbolRefs. Note that any method that works on a SymtabEntry must also work with a SymbolRef, see the @derefmethod macro for a convenient helper macro to generate SymbolRef -> SymtabEntry wrapper methods. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

SymtabEntry()

Util:

deref()

Properties:

symbol_name()

symbol_value()

isundef()

isglobal()

islocal()

isweak()

ObjectFile.__init__

ObjectFile.canonical_rpaths — Function.

canonical_rpaths(rpath::RPath)

Return a canonicalized list of paths that will be searched.

ObjectFile.contents — Function.

contents(oh::ObjectHandle, section::Section)

Read the contents of the section referred to by section from the given ObjectHandle, returning a Vector{UInt8}.

ObjectFile.deref — Function.

deref(section::SectionRef)

Dereference the given SectionRef object to a Section.

deref(seg::SegmentRef)

Dereference the given SegmentRef object to a Segment.

ObjectFile.dylib_compatibility

ObjectFile.MachO.dylib_name — Function.

dylib_name(cmd)

Return the name of the dylib referred to by the given Load Command.

ObjectFile.MachO.dylib_timestamp — Function.

dylib_timestamp(cmd)

Return the build timestamp of the dylib referred to by the given Load Command.

ObjectFile.MachO.dylib_version — Function.

dylib_version(cmd)

Return the version of the dylib referred to by the given Load Command.

ObjectFile.ELF.dyn_entry_is_string — Function.

dyn_entry_is_string(d::ELFDynEntry)

Return true if the given ELFDynEntry represents an offset within the dynamic string table, and therefore can be used in a strtab_lookup()

ObjectFile.ELF.dyn_entry_type — Function.

dyn_entry_type(d::ELFDynEntry)

Return the type of the given ELFDynEntry

ObjectFile.ELF.dyn_entry_type_string — Function.

dyn_entry_type_string(d::ELFDynEntry)

Return the given ELFDynEntry's type as a string.

ObjectFile.eltype

ObjectFile.endianness — Function.

endianness(oh::ObjectHandle)

Returns the endianness of the given ObjectHandle (e.g. :LittleEndian)

Base.eof — Function.

eof(section::SectionRef)

Returns true if the ObjectHandle that this SectionRef refers to has read beyond the current section's extent

ObjectFile.eval

ObjectFile.find_libraries — Function.

find_libraries(oh::ObjectHandle)

Return a mapping from sonames to absolute paths, containing all the sonames declared as beeing needed by the given ObjectHandle. See the documentation for find_library(::RPath, ::String) and RPath for more details.

ObjectFile.find_library — Function.

find_library(oh::ObjectHandle, soname::String)

Return the absolute path to the given soname, using the linker search path that the given ObjectHandle would use at runtime. See the documentation for find_library(::RPath, ::String) for more details.

find_library(rpath::RPath, soname::String)

Return the full path to a library, searching the given RPath, and then the default library search paths. This method takes the given soname and joins it to the end of every path within the given RPath, returning the resultant path if it exists, returning back the original soname if it doesn't.

Base.findall — Function.

findall(sections::Sections, name::String)

Return a list of sections that match the given name.

findall(sections::Sections, name::String)

Return a list of sections that match one of the given names.

Base.findfirst — Function.

findfirst(sections::Sections, name::String)

Return the first section that matches the given name.

findfirst(sections::Sections, names::Vector{String})

Return the first section that matches on of the given names.

ObjectFile.fixed_string — Type.

fixed_string{T}

A fixed-byte string, stored as an integer type (e.g. T = UInt128, or T = UInt64) but displayed and treated as a string.

source

ObjectFile.format_string — Function.

format_string(::Type{H}) where {H <: ObjectHandle}

Return the string name of the given ObjectHandle, examples are "ELF", "MachO", "COFF", etc...

ObjectFile.getindex

ObjectFile.getindex_ref — Function.

getindex_ref(collection, offset, stride, T, ref_type, idx)

Given a collection, such as Sections, DynamicLinks, etc... use the given offset, stride, and T parameters to read in and construct a ref_type object located at index idx. Example invocation:

getindex_ref(
    sections,
    section_header_offset(oh),
    section_header_size(oh),
    section_header_type(oh),
    SectionRef,
    idx
)

ObjectFile.handle — Function.

handle(sections::Sections)

Return the ObjectHandle that this Sections object belongs to

handle(section::SectionRef)

Return the ObjectHandle this SectionRef belongs to. This method is SectionRef-only.

handle(seg::SegmentRef)

Return the ObjectHandle this SegmentRef belongs to. This method is SegmentRef-only.

handle(s::StrTab)

Return the ObjectHandle this StrTab belongs to.

handle(sym::SymbolRef)

Return the ObjectHandle that this SymbolRef belongs to.

handle(rpath::RPath)

Return the handle that this RPath object refers to.

ObjectFile.header — Function.

header(oh::ObjectHandle)

Return the ObjectHandle's header object, whatever that may be for this particular object file format.

ObjectFile.include

ObjectFile.iostream — Function.

iostream(oh::ObjectHandle)

Returns the IOStream backing the ObjectHandle

ObjectFile.is64bit — Function.

is64bit(oh::ObjectHandle)

Returns true if the given ObjectHandle represents a 64-bit object

ObjectFile.isdynamic — Function.

isdynamic(oh::ObjectHandle)

Returns true if the given ObjectHandle makes use of dynamic linking.

ObjectFile.isexecutable — Function.

isexecutable(oh::ObjectHandle)

Returns true if the given ObjectHandle represents an executable object

ObjectFile.isglobal — Function.

isglobal(sym::SymtabEntry)

Return true if the given symbol is global

ObjectFile.islibrary — Function.

islibrary(oh::ObjectHandle)

Returns true if the given ObjectHandle represents a shared library

ObjectFile.islocal — Function.

islocal(sym::SymtabEntry)

Return true if the given symbol is local

ObjectFile.isrelocatable — Function.

isrelocatable(oh::ObjectHandle)

Returns true if the given ObjectHandle represents a relocatable object file, e.g. an .o file as generated by gcc -c

ObjectFile.isundef — Function.

isundef(sym::SymtabEntry)

Return true if the given symbol is undefined

ObjectFile.isweak — Function.

isweak(sym::SymtabEntry)

Return true if the given symbol is weak

ObjectFile.iterate

ObjectFile.lastindex

ObjectFile.length

ObjectFile.mangle_section_name — Function.

mangle_section_name(oh::ObjectHandle, name::AbstractString)

Turn a section name into the object-format specific naming convention, e.g. returning ".bss" for ELF/COFF files, and "__bss" for MachO files

ObjectFile.mangle_symbol_name — Function.

mangle_symbol_name(oh::ObjectHandle, name::AbstractString)

Mangle a symbol name using the object-format specific naming convention, e.g. prefixing "_" for MachO files.

ObjectFile.path — Function.

path(oh::ObjectHandle)

Return the absolute path to the given ObjectHandle, if it was a file loaded from the local disk. If it was loaded from a general IOStream or in some other way such that the path is unknown or unknowable, return the empty string.

path(io::IO)

Try to guess the path of an IO object. If it cannot be guessed, returns the empty string.

ObjectFile.position

Base.read — Function.

read(section::SectionRef)

Read the contents of the section referred to by section, returning a Vector{UInt8}.

ObjectFile.read_struct — Function.

read_struct(oh, type_func, size_func, name)

Given a Type, (such as ELFSection64), unpack() it from the given object and return it, throwing errors as appropriate, and skipping over any excess padding bytes as determined by type_func and size_func. Example invocation:

readstruct(oh, symtabentrytype, symtabentry_size, "Symbol Entry")

ObjectFile.readbytes

ObjectFile.readmeta — Function.

readmeta(io::IO, ::ObjectHandle)

Read an Object File out from an IOStream. This is the first method you should call in order to manipulate object files.

readmeta(io::IO)

Read an Object File out from an IOStream, guessing at the type of object within the stream by calling readmeta(io, T) for each T within ObjTypes, and returning the first that does not throw a MagicMismatch.

readmeta(f::Function, file::AbstractString)

Do-block variant of readmeta(). Use via something like:

readmeta("libfoo.so") do f
    ...
end

ObjectFile.readuntil

ObjectFile.rpaths — Function.

rpaths(rpath::RPath)

Return the list of paths that will be searched for shared libraries.

ObjectFile.section_address — Function.

section_address(section::Section)

The address of the section in virtual memory.

ObjectFile.section_header_offset — Function.

section_header_offset(oh::ObjectHandle)

Given an ObjectHandle, return the offset (in bytes) at which the sections start within the containing object file.

ObjectFile.section_header_size — Function.

section_header_size(oh::ObjectHandle)

Given an ObjectHandle, return the size of a section header (used for reading in the sections header when trying to load a Section object or iterating over a Sections object)

ObjectFile.section_header_type — Function.

section_header_type(oh::ObjectHandle)

Given an ObjectHandle, return the type of a section header (used for reading in the sections header when trying to load a Section object or iterating over a Sections object). For instance, for a 64-bit ELF file, this would return the type ELFSection64

ObjectFile.section_name — Function.

section_name(section::Section)

Return the name of the given section as a string. In order to return a true name, it is necessary to perform a lookup within the object's string table, which cannot be done using just a Section object; use a SectionRef object instead if you need that. For sanity sake, this method will return a string, but the contents of the string may be something like the offset within the string table pointing to this Section's name, e.g. "@strtab.123"

ObjectFile.section_number — Function.

section_number(section::SectionRef)

The index of the given section within the section header table.

ObjectFile.section_offset — Function.

section_offset(section::Section)

The offset of the section in the file, in bytes

ObjectFile.section_size — Function.

section_size(section::Section)

The size of the actual data contained in the section. This should exclude any padding mandated by the file format e.g. due to alignment rules

ObjectFile.section_type

Base.seek — Function.

seek(section::SectionRef, offset)

Seek to offset relative to section in the ObjectHandle that this SectionRef refers to

Base.seekstart — Function.

seekstart(section::SectionRef)

Seek to the beginning of section in the ObjectHandle it was loaded from.

ObjectFile.segment_address

ObjectFile.segment_file_size

ObjectFile.segment_header_offset — Function.

segment_header_offset(oh::ELFHandle)

Return the offset of the segment header table within the given ELF object.

segment_header_offset(oh::ObjectHandle)

Given an ObjectHandle, return the offset (in bytes) at which the segments start within the containing object file.

ObjectFile.segment_header_size — Function.

segment_header_offset(oh::ELFHandle)

Return the size of a segment header within the given ELF object.

segment_header_size(oh::ObjectHandle)

Given an ObjectHandle, return the size of a segment header (used for reading in the segments header when trying to load a Segment object or iterating over a Segments object)

ObjectFile.segment_header_type — Function.

segment_header_type(oh::ELFHandle)

Return the type of a segment header within the given ELF object. E.g. within a 64-bit ELF object, this will return ELFSegment64.

segment_header_type(oh::ObjectHandle)

Given an ObjectHandle, return the type of a segment header (used for reading in the segments header when trying to load a Segment object or iterating over a Segments object). For instance, for a 64-bit ELF file, this would return the type ELFSegment64

ObjectFile.segment_memory_size

ObjectFile.segment_name — Function.

segment_name(seg::SegmentRef)

The name of the given Segment, returned as a string. This method often performs some kind of lookup within the string table of the object to get the full name of the segment.

ObjectFile.segment_num_sections

ObjectFile.segment_number — Function.

segment_number(seg::SegmentRef)

Return the index of the referred segment.

ObjectFile.segment_offset

ObjectFile.segment_type

ObjectFile.show_collection — Function.

show_collection(io, collection, ::Type{H<:ObjectHandle})

Given a collection-like object, (Symbols, DynamicLinks, ``)

ObjectFile.skip

ObjectFile.startaddr — Function.

startaddr(oh::ObjectHandle)

Returns the offset within the underlying IOStream at which this ObjectHandle is located.

ObjectFile.strtab_lookup — Function.

strtab_lookup(s::StrTab, index)

Reads a string from the given StrTab at index.

ObjectFile.symbol_name — Function.

symbol_name(sym::SymtabEntry)

Return the name of the given section as a string. In order to return a true name, it is necessary to perform a lookup within the object's string table, which cannot be done using just a SymtabEntry object; use a SymbolRef object instead if you need that. For sanity sake, this method will return a string, but the contents of the string may be something like the offset within the string table pointing to this SymtabEntry's name, e.g. "@strtab.123"

symbol_name(sym::SymbolRef)

Return the name of the given symbol as a string. This method often performs some kind of lookup within the string table of the object to get the full name of the symbol.

ObjectFile.symbol_number — Function.

symbol_number(sym::SymbolRef)

Return the number (index) of the given symbol.

ObjectFile.symbol_value — Function.

symbol_value(sym::SymtabEntry)

Return the value of the given symbol

ObjectFile.symtab_entry_offset — Function.

symtab_entry_offset(oh::ObjectHandle)

Given an ObjectHandle, return the offset (in bytes) at which the symbol table starts within the containing object file.

ObjectFile.symtab_entry_size — Function.

symtab_entry_size(oh::ObjectHandle)

Given an ObjectHandle, return the size of a symbol table entry (used for reading in the symbol table when trying to load a SymtabEntry object or iterating over a Symbols object).

ObjectFile.symtab_entry_type — Function.

symtab_entry_type(oh::ObjectHandle)

Given an ObjectHandle, return the type of a symbol table entry (used for reading in the symbol table when trying to load a SymtabEntry object or iterating over a Symbols object). For instance, for a 64-bit ELF file, this would return the type ELFSymtabEntry64

ObjectFile.write