Docstrings

IntervalSets...

IntervalSets.AbstractInterval — Type.

A subtype of AbstractInterval{T} represents an interval subset of type T, that provides endpoints, closedendpoints.

IntervalSets.ClosedInterval — Type.

A ClosedInterval(left, right) is an interval set that includes both its upper and lower bounds. In mathematical notation, the constructed range is [left, right].

IntervalSets.Domain — Type.

A subtype of Domain{T} represents a subset of type T, that provides in.

IntervalSets.Interval — Type.

An Interval{L,R}(left, right) where L,R are :open or :closed is an interval set containg x such that

left ≤ x ≤ right if L == R == :closed

left < x ≤ right if L == :open and R == :closed

left ≤ x < right if L == :closed and R == :open, or

left < x < right if L == R == :open

IntervalSets.IntervalSets

IntervalSets.OpenInterval — Type.

An TypedEndpointsInterval{:open,:open}(left, right) is an interval set that includes both its upper and lower bounds. In mathematical notation, the constructed range is (left, right).

IntervalSets.TypedEndpointsInterval — Type.

A subtype of TypedEndpointsInterval{L,R,T} where L and R are :open or :closed, that represents an interval subset of type T, and provides endpoints.

IntervalSets._checked_conversion

IntervalSets._interval_hash

IntervalSets._left_intersect_type

IntervalSets._leq_union

IntervalSets._right_intersect_type

IntervalSets._union

IntervalSets.checked_conversion

IntervalSets.closedendpoints — Function.

A tuple of Bool's encoding whether the left/right endpoints are closed.

IntervalSets.convert

IntervalSets.duration — Function.

duration(iv)

calculates the the total number of integers or dates of an integer or date valued interval. For example, duration(0..1) is 2, while width(0..1) is 1.

IntervalSets.endpoints — Function.

A tuple containing the left and right endpoints of the interval.

IntervalSets.eval

IntervalSets.include

IntervalSets.infimum

IntervalSets.isclosed

IntervalSets.isleftclosed — Function.

Is the interval closed at the left endpoint?

IntervalSets.isleftopen — Function.

Is the interval open at the left endpoint?

IntervalSets.isopen

IntervalSets.isrightclosed — Function.

Is the interval closed at the right endpoint?

IntervalSets.isrightopen — Function.

Is the interval open at the right endpoint?

IntervalSets.leftendpoint — Function.

The left endpoint of the interval.

IntervalSets.length

IntervalSets.ordered

IntervalSets.rightendpoint — Function.

The right endpoint of the interval.

IntervalSets.supremum

IntervalSets.width — Function.

w = width(iv)

Calculate the width (max-min) of interval iv. Note that for integers l and r, width(l..r) = length(l:r) - 1.

IntervalSets.±

IntervalSets.⊇