Docstrings

IntervalContractors.IntervalContractors

IntervalContractors.abs_rev — Function.

Reverse abs

IntervalContractors.acos_rev — Function.

Reverse acos.

IntervalContractors.acosh_rev — Function.

Reverse function for acosh.

IntervalContractors.asin!

IntervalContractors.asin_rev — Function.

Reverse asin.

IntervalContractors.asinh_rev — Function.

Reverse function for asinh.

IntervalContractors.atan_rev — Function.

Reverse atan.

IntervalContractors.atanh_rev — Function.

Reverse function for atanh.

IntervalContractors.constant_contractor

IntervalContractors.cos! — Function.

cos!(X::IntervalBox)

Contractor for cos. Takes an IntervalBox containing the x and y component intervals. Returns an IntervalBox contracted down to the set $y = \cos(x)$.

IntervalContractors.cos_main — Function.

Contractor for main branch of cos, from x = 0 to π.

IntervalContractors.cos_rev — Function.

cos_rev(y::Interval, x::Interval)

Reverse function for cos:

find the subset of x such that $y = \cos(x)$ for the given y.

IntervalContractors.cos_reverse

IntervalContractors.cosh_rev — Function.

Reverse function for cosh.

IntervalContractors.cube!

IntervalContractors.cube_neg

IntervalContractors.cube_pos

IntervalContractors.div_rev — Function.

Reverse division

IntervalContractors.eval

IntervalContractors.exp!

IntervalContractors.exp10_rev — Function.

Reverse function for exp10.

IntervalContractors.exp2_rev — Function.

Reverse function for exp2.

IntervalContractors.exp_rev — Function.

Reverse function for exp.

IntervalContractors.expm1_rev — Function.

Reverse function for expm1.

IntervalContractors.half_pi

IntervalContractors.include

IntervalContractors.integer_contractor

IntervalContractors.inv_rev — Function.

Reverse inverse

IntervalContractors.log!

IntervalContractors.log10_rev — Function.

Reverse function for log10: $y = \log10(x)$

IntervalContractors.log1p_rev — Function.

Reverse function for log1p: $y = \log1p(x)$

IntervalContractors.log2_rev — Function.

Reverse function for log2: $y = \log2(x)$

IntervalContractors.log_rev — Function.

Reverse function for log: $y = \log(x)$

IntervalContractors.max_rev — Function.

Reverse max

IntervalContractors.min_rev — Function.

Reverse min

IntervalContractors.minus_rev — Function.

Reverse minus

IntervalContractors.mul_rev — Function.

Reverse multiplication

IntervalContractors.mul_rev_IEEE1788

IntervalContractors.odd

IntervalContractors.periodise — Function.

Periodize the contractor C

IntervalContractors.plus_rev — Function.

Reverse plus

IntervalContractors.pow_rev1

IntervalContractors.pow_rev2

IntervalContractors.power_rev — Function.

Reverse power

IntervalContractors.reflect_x — Function.

Reflect in mirror at position x_mirror

IntervalContractors.reverse_operations — Constant.

Dictionary mapping functions to their reverse functions.

IntervalContractors.sign_rev

IntervalContractors.sin! — Function.

sin!(X::IntervalBox)

Contractor for sin. Takes an IntervalBox containing the x and y component intervals. Returns an IntervalBox contracted down to the set $y = \sin(x)$.

IntervalContractors.sin_main — Function.

Contractor for "main branch" of sin, from x = -π/2 to π/2.

IntervalContractors.sin_rev — Function.

sin_rev(y::Interval, x::Interval)

Reverse function for sin:

find the subset of x such that $y = \sin(x)$ for the given y.

IntervalContractors.sin_reverse

IntervalContractors.sinh_rev — Function.

Reverse function for sinh.

IntervalContractors.sqr_rev — Function.

Reverse sqr

IntervalContractors.sqrt_rev — Function.

Reverse square root

IntervalContractors.square!

IntervalContractors.square_neg

IntervalContractors.square_pos — Function.

Contractor for y = x^2, x >= 0

IntervalContractors.symmetrise — Function.

Symmetric part of a Contractor, via an involution op (i.e. such that inv(op) == op).

IntervalContractors.tan!

IntervalContractors.tan_main — Function.

Contractor for "main branch" of tan, from x = -π/2 to π/2.

IntervalContractors.tan_rev — Function.

tan_rev(y::Interval, x::Interval)

Reverse function for tan:

find the subset of x such that $y = \tan(x)$ for the given y.

IntervalContractors.tanh_rev — Function.

Reverse function for tanh.

IntervalContractors.translate — Function.

translate(α)

Returns a function that shifts (translates) a 2D IntervalBox in x (the first coordinate).

Translation of a Contractor C by α. Uses inv(op) ∘ C ∘ op

IntervalContractors.two_pi