Docstrings

TreeView.@dag — Macro.

Make a Directed Acyclic Graph (DAG) from a Julia expression.

TreeView.@dag_cse — Macro.

Perform common subexpression elimination on a Julia Expression, and make a Directed Acyclic Graph (DAG) of the result.

TreeView.@tree — Macro.

Make a tree from a Julia Expression. Omits call.

TreeView.@tree_with_call — Macro.

Make a tree from a Julia Expression. Includes call.

TreeView.DirectedAcyclicGraph — Type.

Structure representing a DAG. Maintains a symbol_map giving the currently-known symbols and the corresponding vertex number in the graph.

TreeView.LabelledDiGraph

TreeView.LabelledTree

TreeView.TreeView

TreeView.add_numbered_vertex!

TreeView.add_symbol! — Function.

Adds a symbol to the DAG if it doesn't already exist. Returns the vertex number

Make numbers unique:

TreeView.draw

TreeView.eval

TreeView.include

TreeView.latex_escape

TreeView.latex_label — Function.

Convert a symbol or into a LaTeX label

TreeView.lookup! — Function.

Look up a symbol to see if it has already been seen.

TreeView.make_dag — Function.

Make a Directed Acyclic Graph (DAG) from a Julia expression.

TreeView.make_dag! — Function.

Update a Directed Acyclic Graph with the result of traversing the given Expression.

TreeView.tikz_representation — Function.

Return a Tikz representation of a tree object. The tree object must have fields g (the graph) and labels.

TreeView.walk_tree

TreeView.walk_tree! — Function.

walk_tree!(g, labels, ex, show_call=true)

Walk the abstract syntax tree (AST) of the given expression ex. Builds up the graph g and the set of labels for each node, both modified in place

show_call specifies whether to include call nodes in the graph. Including them represents the Julia AST more precisely, but adds visual noise.

Returns the number of the top vertex.