Docstrings

UnicodePlots.AsciiCanvas — Type.

As the name suggests the AsciiCanvas only uses ASCII characters to draw it's content. Naturally, it doesn't look quite as nice as the Unicode-based ones. However, in some situations it might yield better results. Printing plots to a file is one of those situations.

The AsciiCanvas is best utilized in combination with lineplot. For scatterplot we suggest to use the DotCanvas instead.

UnicodePlots.BarplotGraphics

UnicodePlots.BlockCanvas — Type.

The BlockCanvas is also Unicode-based. It has half the resolution of the BrailleCanvas. In contrast to BrailleCanvas, the pixels don't have visible spacing between them. This canvas effectively turns every character into 4 pixels that can individually be manipulated using binary operations.

UnicodePlots.BrailleCanvas — Type.

The type of canvas with the highest resolution for Unicode-based plotting. It uses the Unicode characters for the Braille symbols to represent individual pixel. This effectively turns every character into 8 pixels that can individually be manipulated using binary operations.

UnicodePlots.Canvas

UnicodePlots.CreateLookupCanvas

UnicodePlots.DensityCanvas — Type.

Unlike the BrailleCanvas, the density canvas does not simply mark a "pixel" as set. Instead it increments a counter per character that keeps track of the frequency of pixels drawn in that character. Together with a variable that keeps track of the maximum frequency, the canvas can thus draw the density of datapoints.

UnicodePlots.DotCanvas — Type.

Similar to the AsciiCanvas, the DotCanvas only uses ASCII characters to draw it's content. Naturally, it doesn't look quite as nice as the Unicode-based ones. However, in some situations it might yield better results. Printing plots to a file is one of those situations.

The DotCanvas is best utilized in combination with scatterplot. For lineplot we suggest to use the AsciiCanvas instead.

UnicodePlots.GraphicsArea

UnicodePlots.LookupCanvas

UnicodePlots.Plot — Type.

Plot(graphics; nargs...) → Plot

Description

Decoration for objects that are GraphicsArea (or Canvas). It is used to surround the inner GraphicsArea object with additional information such as a title, border, and axis labels.

Usage

Plot(graphics; title = "", xlabel = "", ylabel = "", border = :solid, margin = 3, padding = 1, labels = true)

Plot(x, y, canvastype; title = "", width = 40, height = 15, border = :solid, xlim = [0, 0], ylim = [0, 0], margin = 3, padding = 1, labels = true, grid = true)

Arguments

graphics : The GraphicsArea (e.g. a subtype of Canvas) that the plot should decorate

x : The horizontal dimension for each point.

y : The vertical dimension for each point.

canvastype : The type of canvas that should be used for drawing.

title : Text to display on the top of the plot.

xlabel : Text to display on the x axis of the plot

ylabel : Text to display on the y axis of the plot

width : Number of characters per row that should be used for plotting.

height : Number of rows that should be used for plotting. Not applicable to barplot.

border : The style of the bounding box of the plot. Supports :solid, :bold, :dashed, :dotted, :ascii, and :none.

xlim : Plotting range for the x coordinate. [0, 0] stands for automatic.

ylim : Plotting range for the y coordinate. [0, 0] stands for automatic.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

labels : Can be used to hide the labels by setting labels=false.

grid : Can be used to hide the gridlines at the origin

Methods

title!(plot::Plot, title::String)

xlabel!(plot::Plot, xlabel::String)

ylabel!(plot::Plot, xlabel::String)

annotate!(plot::Plot, where::Symbol, value::String)

annotate!(plot::Plot, where::Symbol, row::Int, value::String)

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

see also

scatterplot, lineplot, BarplotGraphics, BrailleCanvas, BlockCanvas, AsciiCanvas

UnicodePlots.UnicodePlots

UnicodePlots._findnz

UnicodePlots.addrow!

UnicodePlots.annotate! — Function.

annotate!(plot, where, value[, color])

annotate!(plot, where, row, value[, color])

This method is responsible for the setting all the textual decorations of a plot.

Note that where can be any of: :tl (top-left), :t (top-center), :tr (top-right), :bl (bottom-left), :b (bottom-center), :br (bottom-right), :l (left), :r (right).

If where is either :l, or :r, then row can be between 1 and the number of character rows of the plots canvas.

UnicodePlots.ascii_decode

UnicodePlots.ascii_lookup

UnicodePlots.ascii_signs

UnicodePlots.barplot — Function.

barplot(text, heights; nargs...) → Plot

Description

Draws a horizontal barplot. It uses the first parameter (text) to denote the names for the bars, and the second parameter (heights) as their values. This means that the two vectors have to have the same length. Alternatively, one can specify a barplot using a dictionary. In that case, the keys will be used as the names and the values, which have to be numeric, will be used as the heights of the bars.

Usage

barplot(text, heights; border = :solid, title = "", margin = 3, padding = 1, color = :blue, width = 40, labels = true, symb = "▪")

barplot(dictionary; nargs...)

Arguments

text : The labels/captions of the bars

heights : The values/heights of the bars

dictionary : A dictonary in which the keys will be used as text

and the values will be utilized as heights.

border : The style of the bounding box of the plot.

Supports :solid, :bold, :dashed, :dotted, :ascii, and :none.

title : Text to display on the top of the plot.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :blue, :red, :yellow

width : Number of characters per row that should be used for plotting.

labels : Can be used to hide the labels by setting labels=false.

symb : Specifies the character that should be used to render the bars

Returns

A plot object of type Plot{BarplotGraphics}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Examples

julia> barplot(["Paris", "New York", "Moskau", "Madrid"],
               [2.244, 8.406, 11.92, 3.165],
               title = "Population")

                        Population
         ┌────────────────────────────────────────┐
   Paris │▪▪▪▪▪▪ 2.244                            │
New York │▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪ 8.406           │
  Moskau │▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪▪ 11.92 │
  Madrid │▪▪▪▪▪▪▪▪▪ 3.165                         │
         └────────────────────────────────────────┘

See also

Plot, histogram, BarplotGraphics

UnicodePlots.barplot! — Function.

barplot!(plot, text, heights; nargs) → Plot

Mutating variant of barplot, in which the first parameter (plot) specifies the existing plot to draw on.

See barplot for more information.

UnicodePlots.block_decode

UnicodePlots.block_signs

UnicodePlots.border_ascii

UnicodePlots.border_bold

UnicodePlots.border_dashed

UnicodePlots.border_dotted

UnicodePlots.border_none

UnicodePlots.border_solid

UnicodePlots.bordermap

UnicodePlots.braille_signs

UnicodePlots.ceil_neg_log10

UnicodePlots.color_cycle

UnicodePlots.color_decode

UnicodePlots.color_encode

UnicodePlots.colors

UnicodePlots.compute_stair_lines

UnicodePlots.den_signs

UnicodePlots.densityplot — Function.

densityplot(x, y; nargs...) → Plot

Description

Draws a density plot for the given points. It uses the first parameter x (which should be a vector or range) to denote the horizontal position of each point, and the second parameter y (which should also be a vector or range) as their vertical position. This means that the two vectors have to have the same length.

Usage

densityplot(x, y; title = "", name = "", width = 40, height = 15, border = :solid, xlim = [0, 0], ylim = [0, 0], margin = 3, padding = 1, color = :white, labels = true)

Arguments

x : The horizontal dimension for each point.

y : The vertical dimension for each point.

title : Text to display on the top of the plot.

name : Annotation of the current drawing to displayed on the right

width : Number of characters per row that should be used for plotting.

height : Number of rows that should be used for plotting. Not applicable to barplot.

border : The style of the bounding box of the plot. Supports :solid, :bold, :dashed, :dotted, :ascii, and :none.

xlim : Plotting range for the x coordinate. [0, 0] stands for automatic.

ylim : Plotting range for the y coordinate. [0, 0] stands for automatic.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :blue, :red, :yellow

labels : Can be used to hide the labels by setting labels=false.

Returns

A plot object of type Plot{DensityCanvas}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Examples

julia> densityplot(randn(1000), randn(1000), title = "Density Plot")

                     Density Plot
       ┌────────────────────────────────────────┐
   2.9 │                    ░                   │
       │                ░    ░  ░               │
       │                ░░░  ░ ░ ░      ░       │
       │             ░░░ ░░▒░▒░░░ ▒ ░           │
       │          ░░ ░░ ░▒░░▒▒▒▒░░░░░░░         │
       │         ░░ ░▒░░░░░▓▒▒░▒░▒░▓░▒░░░       │
       │        ░ ░░░░▓▓▒▓▒▒█▒▓▒▒▓▒▒░▒ ░ ░      │
       │           ░ ▒▒▒▒▓▓▒▓▓▓▓▒█▒▒░░░░░░ ░    │
       │     ░   ░░░░░░░▒▒▓░░▓▒█▒▒▓▓▒▒ ░    ░   │
       │           ░ ░░▒▒░░▒░░▒░░░░ ░░  ░ ░     │
       │         ░   ░ ░ ░░░░░░ ░░ ░░ ░         │
       │              ░  ░ ░ ░  ░ ░   ░░░       │
       │                ░░  ░                   │
       │                                        │
  -3.3 │                                        │
       └────────────────────────────────────────┘
       -3.4                                   2.9

See also

Plot, scatterplot, DensityCanvas

UnicodePlots.densityplot!

UnicodePlots.dot_decode

UnicodePlots.dot_signs

UnicodePlots.eval

UnicodePlots.extend_limits

UnicodePlots.float_round_log10

UnicodePlots.grid

UnicodePlots.height

UnicodePlots.histogram — Function.

histogram(v; nargs...) → Plot

Description

Draws a horizontal histogram of the given vector v. If the parameter bins is not specified, then it will be chosen based on sturges rule. Note that histogram is a simply wrapper for barplot, which means that it supports the same parameter arguments.

Usage

histogram(v; bins = sturges(length(v)), symb = "▇", border = :solid, title = "", margin = 3, padding = 1, color = :blue, width = 40, labels = true)

Arguments

v : Vector or range for which the histogram should be computed.

bins : The approximate number of bins that should be used.

border : The style of the bounding box of the plot.

Supports :solid, :bold, :dashed, :dotted, :ascii, and :none.

title : Text to display on the top of the plot.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :blue, :red, :yellow

width : Number of characters per row that should be used for plotting.

labels : Can be used to hide the labels by setting labels=false.

symb : Specifies the character that should be used to render the bars

Returns

A plot object of type Plot{BarplotGraphics}

Author(s)

Iain Dunning (Github: https://github.com/IainNZ)

Christof Stocker (Github: https://github.com/Evizero)

Kenta Sato (Github: https://github.com/bicycle1885)

Examples

julia> histogram(randn(1000), bins = 15, title = "Histogram")

                             Histogram
              ┌────────────────────────────────────────┐
  (-3.5,-3.0] │ 1                                      │
  (-3.0,-2.5] │ 2                                      │
  (-2.5,-2.0] │▇▇▇▇ 22                                 │
  (-2.0,-1.5] │▇▇▇▇▇▇▇▇▇▇ 53                           │
  (-1.5,-1.0] │▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 78                       │
  (-1.0,-0.5] │▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 141          │
   (-0.5,0.0] │▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 191 │
    (0.0,0.5] │▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 193 │
    (0.5,1.0] │▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 150         │
    (1.0,1.5] │▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 101                  │
    (1.5,2.0] │▇▇▇▇▇▇▇▇ 42                             │
    (2.0,2.5] │▇▇▇ 17                                  │
    (2.5,3.0] │▇ 4                                     │
    (3.0,3.5] │▇ 5                                     │
              └────────────────────────────────────────┘

See also

hist, Plot, barplot, BarplotGraphics

UnicodePlots.include

UnicodePlots.lineplot — Function.

lineplot(x, y; nargs...) → Plot

Description

Draws a path through the given points on a new canvas. It uses the first parameter x (which should be a vector or range) to denote the horizontal position of each point, and the second parameter y (which should also be a vector or range) as their vertical position. This means that the two vectors have to have the same length.

Usage

lineplot(x, y; title = "", name = "", width = 40, height = 15, border = :solid, xlim = [0, 0], ylim = [0, 0], margin = 3, padding = 1, color = :auto, labels = true, canvas = BrailleCanvas, grid = true)

lineplot(fun[, start, end]; title = "", name = "", width = 40, height = 15, border = :solid, xlim = [0, 0], ylim = [0, 0], margin = 3, padding = 1, color = :auto, labels = true, canvas = BrailleCanvas, grid = true)

Arguments

x : The horizontal dimension for each point. Can be a Real number or of type TimeType

y : The vertical dimension for each point.

fun : A function f: R -> R that should be drawn from start to end

title : Text to display on the top of the plot.

name : Annotation of the current drawing to displayed on the right

width : Number of characters per row that should be used for plotting.

height : Number of rows that should be used for plotting. Not applicable to barplot.

border : The style of the bounding box of the plot. Supports :solid, :bold, :dashed, :dotted, :ascii, and :none.

xlim : Plotting range for the x coordinate. [0, 0] stands for automatic.

ylim : Plotting range for the y coordinate. [0, 0] stands for automatic.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :blue, :red, :yellow

labels : Can be used to hide the labels by setting labels=false.

canvas : The type of canvas that should be used for drawing.

grid : Can be used to hide the gridlines at the origin

Returns

A plot object of type Plot{T<:Canvas}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Milktrader (Github: https://github.com/milktrader)

Examples

julia> lineplot([1, 2, 7], [9, -6, 8], title = "My Lineplot")

                    My Lineplot
     ┌────────────────────────────────────────┐
   9 │⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     │⢱⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡠⠊│
     │⠈⡆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠔⠁⠀⠀│
     │⠀⢣⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡠⠊⠁⠀⠀⠀⠀│
     │⠀⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠔⠊⠀⠀⠀⠀⠀⠀⠀│
     │⠀⠀⢇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⠔⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     │⠀⠀⠸⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡠⠊⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     │⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠔⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     │⠀⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡠⠒⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     │⠉⠉⠉⠉⡏⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⢉⠭⠋⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠁│
     │⠀⠀⠀⠀⢱⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠔⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     │⠀⠀⠀⠀⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⡠⠊⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     │⠀⠀⠀⠀⠀⢣⠀⠀⠀⠀⠀⢀⠔⠊⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     │⠀⠀⠀⠀⠀⠸⡀⠀⠀⡠⠔⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
  -6 │⠀⠀⠀⠀⠀⠀⢇⡠⠊⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
     └────────────────────────────────────────┘
     1                                        7

See also

Plot, scatter, stairs, BrailleCanvas, BlockCanvas, AsciiCanvas

UnicodePlots.lineplot!

UnicodePlots.lines!

UnicodePlots.lookup_decode

UnicodePlots.lookup_encode

UnicodePlots.ncols

UnicodePlots.next_color!

UnicodePlots.nrows

UnicodePlots.origin

UnicodePlots.origin_x

UnicodePlots.origin_y

UnicodePlots.pixel!

UnicodePlots.pixel_height

UnicodePlots.pixel_size

UnicodePlots.pixel_width

UnicodePlots.plotting_range

UnicodePlots.plotting_range_narrow

UnicodePlots.points!

UnicodePlots.print_border_bottom

UnicodePlots.print_border_top

UnicodePlots.print_color

UnicodePlots.print_title

UnicodePlots.printrow

UnicodePlots.round_down_subtick

UnicodePlots.round_down_tick

UnicodePlots.round_neg_log10

UnicodePlots.round_up_subtick

UnicodePlots.round_up_tick

UnicodePlots.roundable

UnicodePlots.scatterplot — Function.

scatterplot(x, y; nargs...) → Plot

Description

Draws the given points on a new canvas. It uses the first parameter x (which should be a vector or range) to denote the horizontal position of each point, and the second parameter y (which should also be a vector or range) as their vertical position. This means that the two vectors have to have the same length.

Usage

scatterplot(x, y; title = "", name = "", width = 40, height = 15, border = :solid, xlim = [0, 0], ylim = [0, 0], margin = 3, padding = 1, color = :auto, labels = true, canvas = BrailleCanvas, grid = true)

Arguments

x : The horizontal dimension for each point.

y : The vertical dimension for each point.

title : Text to display on the top of the plot.

name : Annotation of the current drawing to displayed on the right

width : Number of characters per row that should be used for plotting.

height : Number of rows that should be used for plotting. Not applicable to barplot.

border : The style of the bounding box of the plot. Supports :solid, :bold, :dashed, :dotted, :ascii, and :none.

xlim : Plotting range for the x coordinate. [0, 0] stands for automatic.

ylim : Plotting range for the y coordinate. [0, 0] stands for automatic.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :blue, :red, :yellow

labels : Can be used to hide the labels by setting labels=false.

canvas : The type of canvas that should be used for drawing.

grid : Can be used to hide the gridlines at the origin

Returns

A plot object of type Plot{T<:Canvas}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Examples

julia> scatterplot(randn(50), randn(50), title = "My Scatterplot", color = :red)

                    My Scatterplot
       ┌────────────────────────────────────────┐
   2.1 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠁⠀⠀⡧⠀⠀⠀⠀⠀⠀⠀⠀⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⡀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠀⢀⠐⠂⠀⠀⠀⠀⠀⡇⠁⠐⠀⠀⠈⠀⠀⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀│
       │⠠⠀⠀⠀⠀⠀⠐⠀⠀⠈⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠐⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀⠄⡇⢀⠀⠀⠀⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠒⠖⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠚⠒⠒⡗⠒⠒⠒⢒⠒⠒⠒⠒⠒⠒⠒⠖⠒⠒⠒⠒⠒⠢│
       │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠀⠀⠀⠀⠠⠀⠁⡇⠀⠀⠀⣀⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠠⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠀⠀⠠⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠁⡇⠀⠀⠀⠀⢀⠀⠄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠀⠄⠀⠀⠀⠀⠄⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠐⠀⠀⠀⠀⠀⠀⠀⠀⠁⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
  -2.2 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
       └────────────────────────────────────────┘
       -2.1                                   1.9

See also

Plot, lineplot, stairs, BrailleCanvas, BlockCanvas, AsciiCanvas

UnicodePlots.scatterplot!

UnicodePlots.spy — Function.

spy(A; nargs...) → Plot

Description

Plots the sparsity pattern for the given matrix A. This means that a scatterplot that resembles the matrix is drawn, in which only the pixel for non-zero elements of the matrix are set.

If the parameters width and height are not explicitly specified, then the function will attempt to preserve the aspect ratio of the matrix, while also attempting to fit the resulting plot withing the bounding box specified by maxwidth and maxheight

Usage

spy(A; maxwidth = 70, maxheight = 40, title = "Sparsity Pattern", width = 0, height = 0, border = :solid, margin = 3, padding = 1, color = :automatic, labels = true, canvas = BrailleCanvas)

Arguments

A : The matrix of interest, which non-zero elements should be drawn

maxwidth : Maximal number of characters per row that should be used for plotting

maxheight : Maximal number of rows that should be used for plotting.

title : Text to display on the top of the plot.

width : Number of characters per row that should be used for plotting. 0 stands for automatic

height : Number of rows that should be used for plotting. 0 stands for automatic

border : The style of the bounding box of the plot. Supports :solid, :bold, :dashed, :dotted, :ascii, and :none.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :blue, :red, :yellow

labels : Can be used to hide the labels by setting labels=false.

canvas : The type of canvas that should be used for drawing.

Returns

A plot object of type Plot{T<:Canvas}

Author(s)

Dominique (Github: https://github.com/dpo)

Christof Stocker (Github: https://github.com/Evizero)

Jake Bolewski (Github: https://github.com/jakebolewski)

Examples

julia> spy(sprandn(50, 120, .05))

                          Sparsity Pattern
    ┌────────────────────────────────────────────────────────────┐
  1 │⢀⠀⠒⠀⢄⠂⠀⡀⠀⠀⠀⠰⠀⠐⠀⠀⠀⠀⠂⠀⠀⠀⠠⠰⠀⠀⠀⡀⠄⠀⠠⠀⠀⠀⠀⠀⡀⠀⠀⠀⡀⠐⠀⠐⠀⠀⡀⠀⠀⠀⠀⠀⠀⠂⡀⠀⡀⠠⠀⠀│ > 0
    │⠀⠀⠀⠀⠀⢀⢀⠀⠀⠀⠀⠐⠄⠀⢀⠂⠀⠠⠀⠉⠀⠀⠀⠀⢐⠀⠀⠀⠀⠀⠠⠐⠄⠄⠀⠀⠀⠐⠐⠈⢁⠀⠀⠀⠀⠀⠀⠀⠠⠀⠐⡐⠀⠄⠀⠀⠈⠀⠀⠠│ < 0
    │⡀⠀⠀⠀⡀⠊⠀⠀⠀⢀⢀⠀⠄⠀⠂⡠⠂⠐⠐⠀⠀⠄⠀⠀⠀⠄⠀⠀⠀⠀⡀⠀⠀⠁⠀⠀⠀⠀⠀⠀⠠⠀⠂⠀⠠⠀⠀⠀⠆⠄⠈⠀⢀⠀⠀⠀⠐⠄⠀⠀│
    │⠀⠀⠀⠀⠀⠀⠀⢠⠀⢀⠀⠀⠀⢀⠀⠄⠀⠀⠀⠂⠀⠀⠀⠇⠀⢀⠀⠀⠁⠁⠀⠄⠄⠀⠀⠁⠁⠀⡈⠀⠁⠄⠀⠀⠄⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠀│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⡐⢀⠀⠀⠀⠐⠂⠁⢀⠠⠀⠀⠀⠀⠀⠁⠀⠀⠀⠠⠀⠀⠀⠀⢀⠀⢄⠀⠀⠀⠐⠀⠀⠈⠀⠀⠐⠀⠄⠂⠠⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
    │⠠⠀⠠⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠠⠈⠀⠀⠀⠀⢀⠀⠀⠄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
    │⠐⠀⠉⠐⠀⠀⠁⠀⠀⠀⠀⠀⠀⠌⠀⠀⠀⠄⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠌⠀⠀⠄⠀⠀⡀⠀⠀⠀⠀⠀⢀⠈⠄⠀⡀⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀│
    │⠂⠀⠀⠀⠀⠀⠀⠠⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠀⡨⠀⠠⠀⠀⠀⠂⠀⠀⠀⠀⠀⠀⠈⡠⠀⡀⠀⠀⠁⠀⠁⠠⠀⠀⠀⠀⠀⠀⠀⠄⡀⠀⠄⡠⠀⠀⠀⠀│
    │⠐⠀⠀⠀⠀⠀⠀⠃⠀⠀⠀⠀⠁⠀⡀⠁⠀⢂⠂⠀⠀⠀⠈⠁⠀⠀⠁⠁⠄⠀⠀⠀⠀⠀⡀⠀⠀⠀⠂⠀⠀⠀⠀⠐⠀⠂⠀⠂⠁⠀⠆⠈⠀⠀⠄⠄⠀⠀⠀⠀│
    │⠀⠀⠀⠠⠀⠁⠀⠀⠀⠀⠂⠀⠠⠀⠀⠌⠀⠀⠂⠄⠀⡀⠂⠀⠀⡀⠠⠁⠄⠀⠀⠀⠀⠀⠄⠈⠀⢂⠐⠀⢀⠀⠠⠀⠄⠁⡀⠀⠀⠐⠀⠀⡐⠀⠀⠀⠁⠀⠀⠔│
    │⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠁⠃⡀⡁⠀⠀⠀⢀⠀⠀⠂⠀⠀⠠⠀⠀⠠⠀⠀⠀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠀⠬⠀⡀│
    │⠐⠀⠀⠀⠀⠀⠀⠀⠢⠀⠀⠀⠀⠀⢀⠀⠀⠀⠀⡀⠀⠀⠀⠀⠀⠈⠁⠁⠀⠀⠄⠀⠀⠐⠀⠄⠀⠠⠀⠁⡀⠁⠀⠀⠠⠡⠀⠀⠀⠀⠀⠀⠀⠐⠄⠀⠀⠀⠀⡀│
 50 │⠀⠀⠀⠀⠁⠀⠀⠠⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠠⠂⠀⠀⠀⠀⠀⠀⠐⠀⠈⠂⠀⠀⠀⠀⠀⠀⠀⠀⠠⠈⠀⠀⠀⠀⠀⠀⠂⠀⠄⠀⠀⠀⠂⠁⠀⠀⠀⠂⠀│
    └────────────────────────────────────────────────────────────┘
    1                                                          120
                              nz = 275

See also

Plot, scatterplot, BrailleCanvas, BlockCanvas, DotCanvas

UnicodePlots.stairs — Function.

stairs(x, y; nargs...) → Plot

Description

Draws a staircase plot on a new canvas. It uses the first parameter x (which should be a vector or range) to denote the horizontal position of each point, and the second parameter y (which should also be a vector or range) as their vertical position. This means that the two vectors have to have the same length.

Usage

stairs(x, y; style = :post, title = "", name = "", width = 40, height = 15, border = :solid, xlim = [0, 0], ylim = [0, 0], margin = 3, padding = 1, color = :blue, labels = true, canvas = BrailleCanvas, grid = true)

Arguments

x : The horizontal dimension for each point.

y : The vertical dimension for each point.

style : Specifies where the transition of the stair takes plays. Can be either :pre or :post

title : Text to display on the top of the plot.

name : Annotation of the current drawing to displayed on the right

width : Number of characters per row that should be used for plotting.

height : Number of rows that should be used for plotting. Not applicable to barplot.

border : The style of the bounding box of the plot. Supports :solid, :bold, :dashed, :dotted, :ascii, and :none.

xlim : Plotting range for the x coordinate. [0, 0] stands for automatic.

ylim : Plotting range for the y coordinate. [0, 0] stands for automatic.

margin : Number of empty characters to the left of the whole plot.

padding : Space of the left and right of the plot between the labels and the canvas.

color : Color of the drawing. Can be any of :blue, :red, :yellow

labels : Can be used to hide the labels by setting labels=false.

canvas : The type of canvas that should be used for drawing.

grid : Can be used to hide the gridlines at the origin

Returns

A plot object of type Plot{T<:Canvas}

Author(s)

Christof Stocker (Github: https://github.com/Evizero)

Dominique (Github: https://github.com/dpo)

Examples

julia> stairs([1, 2, 4, 7, 8], [1, 3, 4, 2, 7], style = :post, title = "My Staircase Plot")

                My Staircase Plot
    ┌────────────────────────────────────────┐
  7 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⡄⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⢸⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢸│
    │⠀⠀⠀⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠧⠤⠤⠤⠤⠼│
    │⠀⠀⠀⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
  1 │⣀⣀⣀⣀⣀⣸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│
    └────────────────────────────────────────┘
    1                                        8

See also

Plot, lineplot, scatterplot, BrailleCanvas, BlockCanvas, AsciiCanvas

UnicodePlots.stairs!

UnicodePlots.sturges

UnicodePlots.title — Function.

title(plot) → String

Returns the current title of the given plot. Alternatively, the title can be changed with title!.

UnicodePlots.title! — Function.

title!(plot, newtitle) → Plot

Sets a new title for the given plot. Alternatively, the current title can be queried using title.

UnicodePlots.width

UnicodePlots.x_pixel_per_char

UnicodePlots.xlabel — Function.

xlabel(plot) → String

Returns the current label for the x-axis. Alternatively, the x-label can be changed with xlabel!

UnicodePlots.xlabel! — Function.

xlabel!(plot, newlabel) → Plot

Sets a new x-label for the given plot. Alternatively, the current label can be queried using xlabel

UnicodePlots.y_pixel_per_char

UnicodePlots.ylabel — Function.

ylabel(plot) → String

Returns the current label for the y-axis. Alternatively, the y-label can be changed with ylabel!

UnicodePlots.ylabel! — Function.

ylabel!(plot, newlabel) → Plot

Sets a new y-label for the given plot. Alternatively, the current label can be queried using ylabel