Docstrings

FunctionalTables.ColumnOrdering — Type.

struct ColumnOrdering{K, R}

Ordering specification for a column. K::Symbol is a key for ordering by isless, R::Bool == true reverses that for this key. Internal.

FunctionalTables.ColumnOrderingSpecification — Constant.

Types accepted as valid column ordering specifications in the user API.

FunctionalTables.FunctionalTable — Type.

struct FunctionalTable{C<:NamedTuple, O<:Tuple{Vararg{FunctionalTables.ColumnOrdering,N} where N}}

Internal notes

Use accessors length, colums, and ordering to access the fields, property accessors are forwarded to columns.

The only inner constructor is the one where both the length and the ordering is trusted (and thus unchecked). Outer constructors should first wrap the ordering rule, then compute/verify length.

FunctionalTables.FunctionalTables

FunctionalTables.Key — Type.

Type for keys, used internally.

FunctionalTables.Keys

FunctionalTables.OrderingRule — Type.

struct OrderingRule{R, O<:Tuple{Vararg{FunctionalTables.ColumnOrdering,N} where N}}

Rule for dealing with specified orderings. Verifies key uniqueness.

See VerifyOrdering, TrustOrdering, and TryOrdering.

This type and its methods are internal.

FunctionalTables.Picking

FunctionalTables.RLEVector — Type.

struct RLEVector{C, T, S, anyS}

An RLE encoded vector, using negative lengths for missing values. Use the RLEVector{S}(C, T) constructor for creating an empty one.

When an elemenet in counts is positive, it encodes that many of the corresponding element in data.

Negative counts encode values of type S (has to be a concrete singleton). In this case there is no corresponding value in data, ie data may have fewer elements than counts. Note that 0 values in count are reserved, and currently should not happen.

The flag anyS::Bool is true iff there are any values of type S in object.

An RLEVector is iterable.

FunctionalTables.RepeatRow — Type.

RepeatRow(row)

A row repeated as many times as needed. Can be merged to a FunctionalTable, or instantiated with FunctionalTable(len, repeat_row).

FunctionalTables.RepeatValue — Type.

RepeatValue(value, len)

Equivalent to a vector containing len instances of value. Used internally.

FunctionalTables.SHOWROWS — Constant.

Shows this many values from each column in a FunctionalTable.

FunctionalTables.SINKCONFIG — Constant.

Default sink configuration.

FunctionalTables.SINKVECTORS — Constant.

Sink configuration that collects to vectors.

FunctionalTables.SinkConfig — Type.

struct SinkConfig{useRLE, M}

FunctionalTables.SplitTable — Type.

struct SplitTable{K<:NamedTuple, B<:NamedTuple, O<:Tuple{Vararg{FunctionalTables.ColumnOrdering,N} where N}, T<:FunctionalTable}

Implements by using an iterator.

Internals

Each rows from the underlying FunctionalTable is split into index and rest.

Iterator state is

nothing when the rows of the underlying FunctionalTable have been exhausted,

index, rest, itrstate for the next block, where index and rest are the first

(mismatching) row that has not been pushed to the buffers.

FunctionalTables.TableOrdering — Type.

Shorthand for table ordering. Internal.

FunctionalTables.TrustLength — Type.

struct TrustLength

Wrapper type to indicate that the length should not be checked.

Note

The perfect footgun. Only use when the lengths are known and verified by construction.

FunctionalTables.TrustOrdering — Type.

Accept the specified ordering to hold without any checks (except for verifying that column names are valid).

Note

This can lead to incorrect results, use cautiously. VerifyOrdering is recommended instead as it has little overhead.

FunctionalTables.TryOrdering — Type.

Try the specified ordering, when necesary degrade to a subset of it that holds.

FunctionalTables.VerifyOrdering — Type.

Verify that the specified ordering holds. This is the default ordering rule.

FunctionalTables.Wrapping

FunctionalTables._cmp_table_ordering

FunctionalTables._retained_table_ordering

FunctionalTables._showcolcontents

FunctionalTables.aggregator — Function.

aggregator(f)

Wrap a function so that it maps columns of a FunctionalTable to a table with a single row, columwise, ignoring the index. Returns a closure.

Example

map(aggregator(mean), by(ft, :col))

will calculate means after grouping by :col.

FunctionalTables.append1 — Function.

append1(v, elt)

Append elt to v, allocating a new vector and copying the contents.

Type of new collection is calculated using promote_type.

FunctionalTables.by — Function.

by(ft, splitkeys)

An iterator that groups rows of tables by the columns splitkeys, returning (index::NamedTupe, table::FunctionalTable) for each contiguous block of the index keys.

The function has a convenience form by(ft, splitkeys...; ...).

FunctionalTables.cancontain — Function.

cancontain(T, elt)

Test if a collection of element type T can contain a new element elt without any loss of precision.

FunctionalTables.cmp_ordering — Function.

cmp_ordering(ordering, a, b)

Compare a and b given the ordering.

Internal.

FunctionalTables.collect_column — Function.

collect_column(cfg, itr)

Collect results from itr into a sink (using config cfg), then finalize and return the column.

FunctionalTables.collect_column!

FunctionalTables.collect_columns — Function.

len, columns, ordering_rule =

collect_columns(cfg, itr, ordering_rule)
collect_columns(cfg, itr, ordering_rule, known_types)

Collect results from itr, which are supposed to be NamedTuples with the same names, into sinks (using config cfg), then finalize and return

the length,

the NamedTuple of the columns, and

the ordering rule (which is always ::TrustOrdering, by construction).

Determine the names and types from the first named tuple, using known_types as the narrowest types for the given columns.

Special rules for empty iterators

When itr is empty, use a known_types will be used to create empty columns, and only the TryOrdering rule will be narrowed to these. Other rule with more column names may cause an error in the callee, which is intentional.

FunctionalTables.collect_columns!

FunctionalTables.columns — Function.

columns(ft)

Return the columns in a NamedTuple.

Each column is an iterable, but not necessarily an <: AbstractVector.

Note

Never mutate columns obtained by this method, as that will violate invariants assumed by the implementation. Use map(collect, columns(ft)) or similar to obtain mutable vectors.

FunctionalTables.empty_sinks — Function.

empty_sinks(cfg, ?)

Empty sinks for a named tuple of elements, using a type.

FunctionalTables.eval

FunctionalTables.finalize_sink — Function.

finalize_sink(?, sink)

Convert sink to a column.

sink may share structure with the result and is not supposed to be used for saving any more elements.

FunctionalTables.finalize_sinks — Function.

finalize_sinks(cfg, sinks)

Finalize a (named) tuple of sinks.

FunctionalTables.fuse — Function.

fuse(f, index, tables)

Prepend the index as repeated columns to f(index, tables...).

FunctionalTables.ignoreindex — Function.

ignoreindex(f)

Wrap a function returning a closure that ignores the first argument (the index in by mappings).

FunctionalTables.include

FunctionalTables.is_prefix — Function.

is_prefix(a, b)

Test if b starts with a.

FunctionalTables.isless_ordering

FunctionalTables.key_in — Function.

key_in(key, keys)

Check if a key is in itr.

FunctionalTables.key_issubset

FunctionalTables.key_setdiff

FunctionalTables.make_sink — Function.

make_sink(cfg, ?)

Create and return a sink using configuration cfg that stores elements of type T. When T is unkown, use Base.Bottom.

FunctionalTables.mask_ordering — Function.

mask_ordering(ordering, keys)
mask_ordering(ordering, keys, invert)

When invert == false, keep the initial part of ordering that has keys in keys. Not having a key in keys invalidates the tail ordering from that point. This is useful for selecting subsets of orderings.

When invert == true, having a key in keys invalidates the ordering. This is useful for orderings of merged and dropped columns.

FunctionalTables.mask_try_ordering — Function.

mask_try_ordering(ordering_rule, keys)

For TryOrdering, return a masked ordering rule that is is defined on keys so that comparisons make sense, otherwise return the original ordering_rule (that will just error for undefined keys).

FunctionalTables.merge_default_types — Function.

merge_default_types(rowtype, defaults)

Whenever defaults has a given key, use the corresponding type in rowtype, otherwise Union{}.

FunctionalTables.narrow — Function.

narrow(x)

Convert the argument to a narrower type if possible without losing precision.

!!! NOTE This function is not type stable, use only when new container types are determined.

FunctionalTables.ordering — Function.

ordering(ft)

Return the ordering of the table, which is a tuple of ColumnOrdering objects.

FunctionalTables.ordering_repr — Function.

ordering_repr(ordering)

A string representation of an ordering, eg for use in show.

FunctionalTables.orderkey — Function.

Accessor for key. Internal.

FunctionalTables.orderrev

FunctionalTables.picking — Function.

picking(key)

FunctionalTables.rename — Function.

rename(ft, changes)

Rename the columns of a FunctionalTable. The second argument should be a NamedTuple of src = dest pairs, where dest is a symbol.

Example

rename(ft, (α = :a, β = :b)) # rename `α` to `a` and `β` to `b`

FunctionalTables.retained_ordering — Function.

retained_ordering(ordering, a, b)

Return the (initial) part of ordering under which a ≤ b.

FunctionalTables.sort_in_memory

FunctionalTables.split_compatible_ordering — Function.

split_compatible_ordering(ordering, splitkeys)

Extend ordering so that it can be used to split a table with splitkeys.

FunctionalTables.split_namedtuple — Function.

split_namedtuple(?, nt)

Splits a named tuple in two, based on the names in splitter.

Returns two NamedTuples; the first one is ordered as splitter, the second one with the remaining values as in the original argument.

julia> split_namedtuple(NamedTuple{(:a, :c)}, (c = 1, b = 2, a = 3, d = 4))
((a = 3, c = 1), (b = 2, d = 4))

FunctionalTables.split_table — Function.

split_table(ft, splitkeys)

Helper function to set up a SplitTable. Internal.

FunctionalTables.start_sinks — Function.

start_sinks(cfg, row, known_types)

Start sinks using row, using the default known_types when available.

FunctionalTables.store!_or_reallocate — Function.

store!_or_reallocate(?, sink, elt)

Either store elt in sink (in which case the returned value is ≡ sink), or allocate a new sink that can do this, copy the contents, save elt and return that (then the returned value is ≢ sink).

FunctionalTables.store!_or_reallocate_row — Function.

store!_or_reallocate_row(cfg, sinks, elts)

Broadcast store!_or_rellocate for a compatible (named) tuple of sinks and elts. Return the (potentially) new sinks.

FunctionalTables.table_ordering — Function.

table_ordering(column_ordering_specifications)

Process ordering specifications for columns (an iterable or possibly a TableOrdering), return a TableOrdering. Check for uniqueness (but not validity) of keys.

Accepted syntax:

:key, for ordering a column in ascending order,

:key => reverse, for ordering a column in descending order.

All functions which accept ordering specifications should use this, but the function itself is not part of the API.

FunctionalTables.use_rle

FunctionalTables.wrapping — Function.

wrapping(key)

Return a callable that wraps its argument in a NamedTuple with a given key.