Docstrings

StringEncodings.@enc_str

StringEncodings.BUFSIZE

StringEncodings.Encoding

StringEncodings.Encodings

StringEncodings.IConvError

StringEncodings.IncompleteSequenceError

StringEncodings.InvalidEncodingError

StringEncodings.InvalidSequenceError

StringEncodings.OutputBufferError

StringEncodings.StringDecoder — Type.

StringDecoder(stream, from, to=enc"UTF-8")

Returns a new read-only I/O stream, which converts text in the encoding from read from stream into text in the encoding to. Calling close on the stream does not close stream.

to and from can be specified either as a string or as an Encoding object.

Note that some implementations (notably the Windows one) may accept invalid sequences in the input data without raising an error.

StringEncodings.StringEncoder — Type.

StringEncoder(stream, to, from=enc"UTF-8")

Returns a new write-only I/O stream, which converts any text in the encoding from written to it into text in the encoding to written to stream. Calling close on the stream is necessary to complete the encoding (but does not close stream).

to and from can be specified either as a string or as an Encoding object.

StringEncodings.StringEncodingError

StringEncodings.StringEncodings

StringEncodings.__init__

StringEncodings.check_deps

StringEncodings.decode — Function.

decode([T,] a::Vector{UInt8}, enc)

Convert an array of bytes a representing text in encoding enc to a string of type T. By default, a String is returned.

enc can be specified either as a string or as an Encoding object.

StringEncodings.depsjl_path

StringEncodings.encode — Function.

encode(s::AbstractString, enc)

Convert string s to an array of bytes representing text in encoding enc. enc can be specified either as a string or as an Encoding object.

StringEncodings.encoding

StringEncodings.encodings — Function.

encodings()

List all encodings supported by encode, decode, StringEncoder and StringDecoder (i.e. by the current iconv implementation).

Note that encodings typically appear several times under different names. In addition to the encodings returned by this function, the empty string (i.e. "") is equivalent to the encoding of the current locale.

Some implementations may support even more encodings: this can be checked by attempting a conversion. In theory, it is not guaranteed that all conversions between all pairs of encodings are possible; but this is the case with all reasonable implementations.

StringEncodings.encodings_list

StringEncodings.eval

StringEncodings.fill_buffer!

StringEncodings.finalize

StringEncodings.iconv!

StringEncodings.iconv_close

StringEncodings.iconv_close_s

StringEncodings.iconv_open

StringEncodings.iconv_open_s

StringEncodings.iconv_reset!

StringEncodings.iconv_s

StringEncodings.include

StringEncodings.libiconv

StringEncodings.message

StringEncodings.test_encoding

StringEncodings.wrap_stream