Docstrings

WignerSymbols.HalfInteger — Type.

struct HalfInteger <: Real

Represents half-integer values.

HalfInteger(numerator::Integer, denominator::Integer)

Constructs a HalfInteger object as a rational number from the given integer numerator and denominator values.

Examples

julia> HalfInteger(1, 2)
1/2

julia> HalfInteger(-2, 1)
-2

WignerSymbols.HalfIntegerRange — Type.

struct HalfIntegerRange <: AbstractVector{HalfInteger}

A range of HalfInteger values from start to stop, spaced by 1. The a:b syntax where both a and b are HalfIntegers can also be use to construct this range.

WignerSymbols.PrimeFactorization

WignerSymbols.PrimeIterator

WignerSymbols.Wigner3j

WignerSymbols.Wigner6j

WignerSymbols.WignerSymbols

WignerSymbols.__init__

WignerSymbols._vadd!

WignerSymbols._vmax!

WignerSymbols._vmin!

WignerSymbols._vsub!

WignerSymbols.bigone

WignerSymbols.bigprime

WignerSymbols.bigprimetable

WignerSymbols.clebschgordan — Function.

clebschgordan(T::Type{<:AbstractFloat} = Float64, j₁, m₁, j₂, m₂, j₃, m₃ = m₁+m₂) -> ::T

Compute the value of the Clebsch-Gordan coefficient <j₁, m₁; j₂, m₂ | j₃, m₃ > as a type T floating point number. By default, T = Float64 and m₃ = m₁+m₂.

Returns zero(T) if the triangle condition δ(j₁, j₂, j₃) is not satisfied, but throws a DomainError if the jᵢs and mᵢs are not (half)integer or abs(mᵢ) > jᵢ.

WignerSymbols.commondenominator!

WignerSymbols.compute3jseries

WignerSymbols.compute6jseries

WignerSymbols.converthalfinteger

WignerSymbols.divgcd!

WignerSymbols.eval

WignerSymbols.factorialtable

WignerSymbols.factortable

WignerSymbols.include

WignerSymbols.ishalfinteger

WignerSymbols.prime

WignerSymbols.primefactor

WignerSymbols.primefactorial

WignerSymbols.primes

WignerSymbols.primetable

WignerSymbols.racahV — Function.

racahV(T::Type{<:AbstractFloat} = Float64, j₁, j₂, j₃, m₁, m₂, m₃ = -m₁-m₂) -> ::T

Compute the value of Racah's V-symbol V(j₁, j₂, j₃; m₁, m₂, m₃) = (-1)^(-j₁+j₂+j₃) * ⎛ j₁ j₂ j₃ ⎞ ⎝ m₁ m₂ m₃ ⎠ as a type T floating point number. By default, T = Float64 and m₃ = -m₁-m₂.

Returns zero(T) if the triangle condition δ(j₁, j₂, j₃) is not satisfied, but throws a DomainError if the jᵢs and mᵢs are not (half)integer or abs(mᵢ) > jᵢ.

WignerSymbols.racahW — Function.

racahW(T::Type{<:AbstractFloat} = Float64, j₁, j₂, J, j₃, J₁₂, J₂₃) -> ::T

Compute the value of Racah's W coefficient W(j₁, j₂, J, j₃; J₁₂, J₂₃) = <(j₁,(j₂j₃)J₂₃)J | ((j₁j₂)J₁₂,j₃)J> / sqrt((2J₁₂+1)*(2J₁₃+1)) as a type T floating point number. By default, T = Float64.

Returns zero(T) if any of triangle conditions δ(j₁, j₂, J₁₂), δ(j₂, j₃, J₂₃), δ(j₁, J₂₃, J), δ(J₁₂, j₃, J) are not satisfied, but throws a DomainError if the jᵢs and Js are not integer or halfinteger.

WignerSymbols.reorder3j

WignerSymbols.reorder6j

WignerSymbols.splitsquare

WignerSymbols.sumlist!

WignerSymbols.wigner3j — Function.

wigner3j(T::Type{<:AbstractFloat} = Float64, j₁, j₂, j₃, m₁, m₂, m₃ = -m₂-m₁) -> ::T

Compute the value of the Wigner-3j symbol ⎛ j₁ j₂ j₃ ⎞ ⎝ m₁ m₂ m₃ ⎠ as a type T floating point number. By default, T = Float64 and m₃ = -m₁-m₂.

Returns zero(T) if the triangle condition δ(j₁, j₂, j₃) is not satisfied, but throws a DomainError if the jᵢs and mᵢs are not (half)integer or abs(mᵢ) > jᵢ.

WignerSymbols.wigner6j — Function.

wigner6j(T::Type{<:AbstractFloat} = Float64, j₁, j₂, j₃, j₄, j₅, j₆) -> ::T

Compute the value of the Wigner-6j symbol ⎧ j₁ j₂ j₃ ⎫ ⎩ j₄ j₅ j₆ ⎭ as a type T floating point number. By default, T = Float64.

Returns zero(T) if any of triangle conditions δ(j₁, j₂, j₃), δ(j₁, j₆, j₅), δ(j₂, j₄, j₆), δ(j₃, j₄, j₅) are not satisfied, but throws a DomainError if the jᵢs are not integer or halfinteger.

WignerSymbols.Δ — Function.

Δ(T::Type{<:AbstractFloat} = Float64, j₁, j₂, j₃) -> ::T

Computes the triangle coefficient Δ(j₁, j₂, j₃) = √((j₁+j₂-j₃)!*(j₁-j₂+j₃)!*(j₂+j₃-j₁)! / (j₁+j₂+j₃+1)!) as a type T floating point number.

Returns zero(T) if the triangle condition δ(j₁, j₂, j₃) is not satisfied, but throws a DomainError if the jᵢs are not (half)integer

WignerSymbols.Δ²

WignerSymbols.δ — Function.

δ(j₁, j₂, j₃) -> ::Bool

Checks the triangle conditions j₃ <= j₁ + j₂, j₁ <= j₂ + j₃ and j₂ <= j₃ + j₁.

WignerSymbols.ϵ