Docstrings

RomanNumerals.@rn_str — Macro.

rn"MMXV" syntactic sugar

RomanNumerals.NUMERAL_MAP

RomanNumerals.RN — Type.

RomanNumeral(::Integer)
RomanNumeral(::AbstractString)

Roman number: it contains both the standard number and its roman representation. It behaves like standard numbers.

RomanNumerals.RomanNumeral — Type.

RomanNumeral(::Integer)
RomanNumeral(::AbstractString)

Roman number: it contains both the standard number and its roman representation. It behaves like standard numbers.

RomanNumerals.RomanNumerals

RomanNumerals.VALID_ROMAN_PATTERN

RomanNumerals.eval

RomanNumerals.include

RomanNumerals.isprime

RomanNumerals.toroman