Docstrings

COSMO.AbstractConvexCone

COSMO.AbstractConvexSet

COSMO.Affine

COSMO.AffineConvertible

COSMO.Box

COSMO.CI

COSMO.COSMO

COSMO.CompositeConvexSet

COSMO.Constraint — Type.

Constraint(A, b, convex_set, dim = 0, indices = 0:0)

Creates a COSMO constraint: Ax + b ∈ convex_set.

By default the following convex sets are supported: ZeroSet, Nonnegatives, SecondOrderCone, PsdCone.

Examples

julia> Constraint([1 0;0 1], zeros(2), COSMO.PsdCone)
Constraint
Size of A: (2, 2)
ConvexSet: COSMO.PsdCone

The optinal arguments dim and indices can be used to specify A and b for subparts of variable x. If x has dimension dim = 4, then x[2] and x[3] can be constrained to the zero cone in the following way:

Examples

julia> c = Constraint([1 0;0 1], zeros(2), COSMO.ZeroSet, 4, 2:3)
Constraint
Size of A: (2, 4)
ConvexSet: COSMO.ZeroSet

Notice that extra columns of A have been added automatically.

julia>Matrix(c.A)
2×4 Array{Float64,2}:
0.0  1.0  0.0  0.0
0.0  0.0  1.0  0.0

COSMO.DefaultFloat

COSMO.DefaultInt

COSMO.EqualTo

COSMO.Flags

COSMO.GreaterThan

COSMO.IdentityMatrix

COSMO.Info

COSMO.Interval

COSMO.IntervalConvertible

COSMO.LessThan

COSMO.LinsolveSubarray

COSMO.MOI

COSMO.MOIU

COSMO.Model

COSMO.Nonnegatives

COSMO.Nonpositives

COSMO.Optimizer

COSMO.PSD

COSMO.PSDS

COSMO.PSDT

COSMO.ProblemData

COSMO.PsdCone

COSMO.PsdConeTriangle

COSMO.Quadratic

COSMO.Result — Type.

Result{T <: AbstractFloat}

Object returned by the COSMO solver after calling optimize!(model, settings). It has the following fields:

Fieldname	Type	Description

x	Vector{T}	Primal variable

y	Vector{T}	Dual variable

s	Vector{T}	(Primal) set variable

obj_val	T	Objective value

iter	Int64	Number of iterations

status	Symbol	Solution status

info	COSMO.ResultInfo	Struct with more information

times	COSMO.ResultTimes	Struct with several measured times

COSMO.ResultInfo

COSMO.ResultTimes

COSMO.SOC

COSMO.ScaleMatrices

COSMO.SecondOrderCone

COSMO.Settings — Type.

Settings(;arg=val)

Creates a COSMO settings object that is used to pass user settings to the solver.

Argument	Description	Values (default)

rho	ADMM rho step	0.1

sigma	ADMM sigma step	1e-6.

alpha	Relaxation parameter	1.6

eps_abs	Absolute residual tolerance	1e-4

eps_rel	Relative residual tolerance	1e-4

epspriminf	Primal infeasibility tolerance	1e-4

epsdualinf	Dual infeasibility tolerance	1e-4

max_iter	Maximum number of iterations	2500

verbose	Verbose printing	false

verbose_timing	Verbose timing	false

check_termination	Check termination interval	40

check_infeasibility	Check infeasibility interval	40

scaling	Number of scaling iterations	10

adaptive_rho	Automatic adaptation of step size parameter	true

time_limit	set solver time limit in s	0

COSMO.SingleVariable

COSMO.SparseTriplets

COSMO.SplitVector

COSMO.SplitView

COSMO.SupportedVectorSets

COSMO.VI

COSMO.Variables

COSMO.VectorAffine

COSMO.VectorOfVariables

COSMO.Workspace — Type.

Model()

Initializes an empty COSMO model that can be filled with problem data using assemble!(model, P, q,constraints, [settings]).

COSMO.ZeroSet

COSMO.Zeros

COSMO._project!

COSMO._scalecoef

COSMO._shift

COSMO._sympackedto

COSMO._unshift

COSMO._warm_start!

COSMO.adapt_rho_vec!

COSMO.admm_step!

COSMO.assemble! — Function.

assemble!(model, P, q, constraint(s), [settings, x0, y0, s0])

Assembles a COSMO.Model with a cost function defind by P and q, and a number of constraints.

The positive semidefinite matrix P and vector q are used to specify the cost function of the optimization problem:

min   1/2 x'Px + q'x
s.t.  Ax + b ∈ C

constraints is a COSMO.Constraint or an array of COSMO.Constraint objects that are used to describe the constraints on x.

The optinal arguments x0, s0, and y0 can be used to provide the solver with warm starting values for the primal variable x, the primal slack variable s and the dual variable y. The optinal argument settings can be used to pass custom solver settings.

COSMO.assign_constraint_row_ranges!

COSMO.calculate_residuals

COSMO.check_dimensions

COSMO.clip

COSMO.coefficient

COSMO.col_norms!

COSMO.constant

COSMO.constraint_rows

COSMO.convert_copy

COSMO.empty!

COSMO.eval

COSMO.extract_upper_triangle!

COSMO.factor_KKT!

COSMO.get_set_indices

COSMO.get_subset

COSMO.has_converged

COSMO.in_dual

COSMO.in_pol_recc

COSMO.include

COSMO.inv_sqrt!

COSMO.is_dual_infeasible

COSMO.is_neg_sem_def

COSMO.is_pos_sem_def

COSMO.is_primal_infeasible

COSMO.kkt_col_norms!

COSMO.limit_scaling

COSMO.limit_scaling!

COSMO.lrmul!

COSMO.max_res_component_norm

COSMO.merge_constraints!

COSMO.merge_nonnegatives

COSMO.merge_zeros

COSMO.num_subsets

COSMO.optimize! — Function.

optimize!(model)

Attempts to solve the optimization problem defined in COSMO.Model object with the user settings defined in COSMO.Settings. Returns a COSMO.Result object.

COSMO.orderval

COSMO.output_index

COSMO.populate_upper_triangle!

COSMO.print_header

COSMO.print_iteration

COSMO.print_result

COSMO.processConstant!

COSMO.processConstraint!

COSMO.processSet!

COSMO.process_constraint!

COSMO.processconstraints

COSMO.processconstraints!

COSMO.processlinearterms!

COSMO.processobjective

COSMO.project!

COSMO.rectify_scalar_scaling!

COSMO.rectify_scaling!

COSMO.rectify_set_scalings!

COSMO.reset!

COSMO.reverse_scaling!

COSMO.row_norms!

COSMO.scale!

COSMO.scale_data!

COSMO.scale_ruiz!

COSMO.scale_sets!

COSMO.scalecoef

COSMO.scaled_norm

COSMO.scaled_norm1

COSMO.scaled_norm2

COSMO.scaled_norm_Inf

COSMO.set! — Function.

set!(model, P, q, A, b, convex_sets, [settings])

Sets model data directly based on provided fields.

COSMO.set_rho_vec!

COSMO.set_user_settings!

COSMO.setup!

COSMO.sort_sets

COSMO.support_function

COSMO.symmetrize!

COSMO.sympackedUtoL

COSMO.sympackeddim

COSMO.sympackedlen

COSMO.trimap

COSMO.trimapL

COSMO.unscalecoef

COSMO.update_rho_vec!

COSMO.variable_index_value

COSMO.warmStart!

COSMO.warm_start_dual!

COSMO.warm_start_primal! — Function.

warmstartprimal!(model, x0, [ind])

Provides the COSMO.Model with warm starting values for the primal variable x. ind can be used to warm start certain components of x.

COSMO.warm_start_slack! — Function.

warmstartslack!(model, s0, [ind])

Provides the COSMO.Model with warm starting values for the primal slack variable s. ind can be used to warm start certain components of s.