Docstrings

RangeArrays.RangeArrays

RangeArrays.RangeMatrix — Type.

RangeMatrix(rs::AbstractVector{T}) where T <: AbstractRange

A RangeMatrix is a simple matrix representation of a vector of ranges, with each range representing one column. Construct a RangeMatrix with a vector of ranges; the ranges must all have the same length.

Note that it is only efficient when all the ranges are of the same type and in a concretely typed Vector.

RangeArrays.RepeatedRangeMatrix — Type.

RepeatedRangeMatrix(r::AbstractRange{T}, at::AbstractVector{T}) where T

A RepeatedRange is like a RangeMatrix, except that it only stores one range and a vector of offsets, at which the range repeats. For now, both the range and vector of offsets must have the same element type.

RangeArrays.eval

RangeArrays.include

RangeArrays.unsafe_getindex