Docstrings

AxisAlgorithms.A_ldiv_B_md — Function.

A_ldiv_B_md(F, src, dim) solves F for slices b of src along dimension dim, storing the result along the same dimension of the output. Currently, F must be an LU-factorized tridiagonal matrix or a Woodbury matrix.

AxisAlgorithms.A_ldiv_B_md! — Function.

A_ldiv_B_md!(dest, F, src, dim) solves a tridiagonal system along dimension dim of src, storing the result in dest. Currently, F must be an LU-factorized tridiagonal matrix. If desired, you may safely use the same array for both src and dest, so that this becomes an in-place algorithm.

AxisAlgorithms.A_mul_B_md — Function.

A_mul_B_md(M, src, dim) computes M*x for slices x of src along dimension dim, storing the resulting vector along the same dimension of the output. M must be an AbstractMatrix. This uses an in-place naive algorithm.

AxisAlgorithms.A_mul_B_md! — Function.

A_mul_B_md!(dest, M, src, dim) computes M*x for slices x of src along dimension dim, storing the result in dest. M must be an AbstractMatrix. This uses an in-place naive algorithm.

AxisAlgorithms.A_mul_B_perm — Function.

A_mul_B_perm(M, src, dim) computes M*x for slices x of src along dimension dim, storing the resulting vector along the same dimension of the output. M must be an AbstractMatrix. This uses permutedims to make dimension dim into the first dimension, performs a standard matrix multiplication, and restores the original dimension ordering. In many cases, this algorithm exhibits the best cache behavior.

AxisAlgorithms.A_mul_B_perm! — Function.

A_mul_B_perm!(dest, M, src, dim) computes M*x for slices x of src along dimension dim, storing the result in dest. M must be an AbstractMatrix. This uses permutedims to make dimension dim into the first dimension, performs a standard matrix multiplication, and restores the original dimension ordering. In many cases, this algorithm exhibits the best cache behavior.

AxisAlgorithms.AxisAlgorithms

AxisAlgorithms._A_ldiv_B_md

AxisAlgorithms._A_ldiv_B_md!

AxisAlgorithms._A_mul_B_md

AxisAlgorithms._A_mul_B_md!

AxisAlgorithms._A_mul_B_md_2x2!

AxisAlgorithms._mul!

AxisAlgorithms._mul2!

AxisAlgorithms.alloc_matmul

AxisAlgorithms.check_matmul_sizes

AxisAlgorithms.check_matrix

AxisAlgorithms.eval

AxisAlgorithms.include

AxisAlgorithms.sub!