Docstrings

ObjectStores.Bucket

ObjectStores.Object

ObjectStores.ObjectStore — Type.

Required fields:

id::String; # From Authorization.AbstractClient

id2permission::Dict{String, Permission}; # Resource ID => Permission (from Authorization.AbstractClient)

idpattern2permission::Dict{Regex, Permission}; # Resource ID pattern => Permission (from Authorization.AbstractClient)

type2permission::Dict{DataType, Permission}; # Resource type => Permission (from Authorization.AbstractClient)

rootbucketID::String # ID of root bucket # Specific to ObjectStores

ObjectStores.ObjectStores

ObjectStores.Permission

ObjectStores.createbucket! — Function.

Create bucket. If successful return nothing, else return an error message as a String.

Base.delete! — Function.

Delete object. If successful return nothing, else return an error message as a String.

ObjectStores.deletebucket! — Function.

Delete bucket. If successful return nothing, else return an error message as a String.

ObjectStores.eval

Base.getindex — Function.

Read object. If successful return the value, else @warn the error message and return nothing.

ObjectStores.getpermission

ObjectStores.haspermission

ObjectStores.include

ObjectStores.isbucket — Function.

Returns true if name refers to a bucket.

ObjectStores.isobject — Function.

Returns true if name refers to an object.

ObjectStores.listcontents — Function.

List the contents of the bucket. If successful return the value, else @warn the error message and return nothing.

ObjectStores.permissions_conflict

ObjectStores.setexpiry!

Base.setindex! — Function.

Create/update object. If successful return nothing, else return an error message as a String.

ObjectStores.setpermission!