Docstrings

KeywordDispatch.@kwdispatch — Macro.

@kwdispatch sig [methods]

Designate a function signature sig that should dispatch on keyword arguments. A function can also be provided, in which case all calls to that function are will dispatch to keyword methods.

It is possible to specify keyword aliases by specifying from => to pairs in the keyword position.

The optional methods argument allows a block of keyword methods specified as anonymous functions. To define additional keyword methods, use the @kwmethod macro.

Examples

@kwdispatch f(_)
@kwdispatch f(_::Real)

@kwdispacth f # equivalent to @kwdispatch f(_...)

@kwdispatch f(x) begin
    (a) -> x+a
    (b) -> x-b
end
# equivalent to
#  @kwdispatch f(x)
#  @kwmethod f(x;a) = x+a
#  @kwmethod f(x;b) = x-b

@kwdispatch f(; alpha=>α) # specifies alpha as an alias to α

KeywordDispatch.@kwmethod — Macro.

@kwmethod expr

Define a keyword method for dispatch. expr should be a standard function definition (either block or inline) with a keyword argument block (which may be empty).

The positional signature should first be designated by the @kwdispatch macro.

Examples

@kwdispatch f() # designate the no positional argument form

@kwmethod f(;) = nothing # no keyword arguments
@kwmethod f(;a) = a
@kwmethod f(;b) = 2*b
@kwmethod f(;a,b) = a+2*b

KeywordDispatch.KeywordDispatch

KeywordDispatch.KeywordMethodError

KeywordDispatch.argmeth

KeywordDispatch.argsym

KeywordDispatch.argtype

KeywordDispatch.eval

KeywordDispatch.generate_kwmethods

KeywordDispatch.include

KeywordDispatch.kwcall

KeywordDispatch.kwmethod_expr

KeywordDispatch.ntsort

KeywordDispatch.unwrap_where

KeywordDispatch.wrap_where