Docstrings

GaloisFields.@GaloisField — Macro.

F = @GaloisField 3
F = @GaloisField ℤ/3ℤ
F = @GaloisField 𝔽₃
F,α = @GaloisField 3^2

Different ways of defining a finite field.

For defining a Galois field of prime power order with $n>1$, consider using @GaloisField! instead, which allows specifying the display name of the generator.

GaloisFields.@GaloisField! — Macro.

const G = @GaloisField! 3 β^2 + 1
const G = @GaloisField! 𝔽₃ β^2 + 1
const G = @GaloisField! 3^2 β
const K = GaloisField(3)
const G = @GaloisField! K β^2 + 1

Define a finite field G and inject a variable for its primitive element into the current scope.

Note that the variable name (e.g. β above) is part of the type. This lets you define identifications between isomorphic (sub)fields. For example, with the following definition

const F = @GaloisField! 𝔽₂ β^2 + β + 1
const G = @GaloisField! 𝔽₂ γ^2 + γ + 1

the fields $F$ and $G$ are isomorphic, but not canonically. We might define

@GaloisFields.identify β => γ + 1
@GaloisFields.identify γ => β + 1

to allow for conversions like

G(β)
convert(F, γ + 1)

GaloisFields.@identify

GaloisFields.AbstractExtensionField — Type.

abstract type AbstractExtensionField <: AbstractGaloisField end

A type representing a finite extension of an underlying finite field.

GaloisFields.AbstractGaloisField — Type.

abstract type AbstractGaloisField <: Number end

A type representing finite fields.

GaloisFields.BITREVERSAL

GaloisFields.BinaryField

GaloisFields.Bits

GaloisFields.BitsIter

GaloisFields.DATAFILE

GaloisFields.Direct — Type.

Direct()

A helper singleton for arithmetic operations that benefit from a log-table. Field implementations implement e.g.

*(::Direct, a, b)

and then define

*(a::F, b::F) = zech_op(F, *, a, b)

This will use the ::Direct implementation for computing a log table where appropriate, and will either use the log table or the ::Direct implementation for computing results.

GaloisFields.ExtensionField — Type.

F = ExtensionField{F <: AbstractGaloisField, N, α, MinPoly}

Algebraic extension of a finite field $F$ of degree $N$.

GaloisFields.GaloisField — Function.

const F = GaloisField(p)
const F,α = GaloisField(p, :β => [1, 0, 1])
const F,α = GaloisField(p, n, :β)

Return a type representing a finite field.

The single-argument signature returns the finite field $ℤ/pℤ$.

The two-arguments signature returns an algebraic extension of that field, with minimum polynomial given by the second argument: a dense representation of the univariate, monic polynomial, with ascending degree.

The three-arguments signature returns an algebraic extension of that field, with minimum polynomial equal to the Conway polynomial for $(p,n)$. The GaloisFields package ships with a database of Conway polynomials and will raise an error if it does not contain an entry for $(p,n)$.

Note that in the latter two cases, the variable name (e.g. β above) is part of the type. This lets you define identifications between isomorphic (sub)fields. For example, with the following definition

const F = @GaloisField! 𝔽₂ β^2 + β + 1
const G = @GaloisField! 𝔽₂ γ^2 + γ + 1

the fields $F$ and $G$ are isomorphic, but not canonically. We might define

@GaloisFields.identify β => γ + 1
@GaloisFields.identify γ => β + 1

to allow for conversions like

G(β)
convert(F, γ + 1)

In the Conway case, you do not have to define your own identifications, as the Conway polynomials satisfy compatibility relations that allow us to use certain distinguished inclusions between them.

GaloisFields.GaloisFields — Module.

GaloisFields.jl - finite fields for Julia

Build Status	Test coverage

[![][travis-img]][travis-url] [![][appveyor-img]][appveyor-url]	[![Coverage Status][codecov-img]][codecov-url]

Introduction

This module defines types representing [finite fields][galois-fields-wiki]. It supports both fields of prime order and of prime power order.

[galois-fields-wiki]: https://en.wikipedia.org/wiki/Finite_field

Synopsis

The easiest way to create Galois fields is with the @GaloisField and @GaloisField! macros. Typically, you use the former for a field of prime order and the latter for a field of prime power order. In the prime power case, you pass a display name / variable name for the primitive element.

using GaloisFields

const F = @GaloisField 29     # ℤ/29ℤ
const G = @GaloisField! 27 β   # degree-3 extension of ℤ/3ℤ; multiplicatively generated by β

F(2)^29 == F(2)
β^27 == β

The macros also accept pretty-printed versions: more difficult to type but more elegant to read:

const F = @GaloisField ℤ/29ℤ
const G = @GaloisField 𝔽₂₇ β

If you want to pass your own generator for the representation of a field of order $q = p^n$, you can:

const F = @GaloisField! 𝔽₃ β^2 + β + 2
β^2 + β + 2 == 0

Lastly, there's also function interfaces in cases where macros are not appropriate:

const F = GaloisField(29)               # ℤ/29ℤ
const G, β = GaloisField(81, :β)        # degree-4 extension of ℤ/3ℤ
const G, β = GaloisField(3, 4, :β)      # same; avoid having to factorize 81
const F, β = GaloisField(3, :β => [2, 0, 0, 2, 1]) # same; pass our own custom minimum polynomial

Fast multiplications

In some cases, we make use of [Zech's logarithms][zech] for faster multiplications. By default, this happens if the order of the field is less than $2^16$, if the characteristic is not 2, and if the primitive element is also a multiplicative generator. However, you can override this by calling either of

GaloisFields.enable_zech_multiplication(F)
GaloisFields.disable_zech_multiplication(F)

before doing any multiplication operation. If you call this function on a field whose primitive element is not a multiplicative generator, this will throw a warning.

[zech]: https://en.wikipedia.org/wiki/Zech's_logarithm

Conversions

If you specify your own minimum polynomial, we make no assumptions about conversions between fields. For example, when defining

const F = @GaloisField! 𝔽₂ β^2 + β + 1
const G = @GaloisField! 𝔽₂ γ^2 + γ + 1

an operation like

G(β)

will throw an error. The mathematical reason is that the fields $F$ and $G$ are isomorphic, but there is two different isomorphisms. ("They are not canonically isomorphic.") To choose an identification, you can use the identify function (which is not exported by default, so we use its full path):

GaloisFields.identify(β => γ^2)
GaloisFields.identify(γ => β^2)

This allows for conversions such as

G(β)
convert(F, γ + 1)

The inner workings of this distinction are based on the symbol names. So if you define $F$ and $G$ with the same symbol and minimum polynomial:

const F = @GaloisField! 𝔽₂ β^2 + β + 1
const G = @GaloisField! 𝔽₂ β^2 + β + 1

then they are just considered equal and conversions work without extra work.

Conversions for the default minimum polynomials

If you do not specify a minimum polynomial, for example by using

const F = @GaloisField! 𝔽₈₁ β
const G = @GaloisField! 𝔽₉ γ

then we use [Conway polynomials][conway]. They have special compatibility relations between them, allowing conversions:

β^10 == γ

This works provided F and G have the same characteristic p. If the order of either is a power of the other, we convert into the bigger field. If not, we convert both into the field of order p^N, where N is the least common multiple of the extension degrees of F and G over ℤ/pℤ.

[conway]: https://en.wikipedia.org/wiki/Conwaypolynomial(finite_fields)

[travis-img]: https://travis-ci.org/tkluck/GaloisFields.jl.svg?branch=master [travis-url]: https://travis-ci.org/tkluck/GaloisFields.jl

[appveyor-img]: https://ci.appveyor.com/api/projects/status/4g6ax1ni7ijx3rn4?svg=true [appveyor-url]: https://ci.appveyor.com/project/tkluck/galoisfields-jl

[codecov-img]: https://codecov.io/gh/tkluck/GaloisFields.jl/branch/master/graph/badge.svg [codecov-url]: https://codecov.io/gh/tkluck/GaloisFields.jl

GaloisFields.InclusionError

GaloisFields.PrimeField — Type.

PrimeField{I<:Integer, p}

A type representing an element in ℤ/pℤ.

GaloisFields.Reduced — Type.

Reduced()

A helper singleton used for asserting that an input value has already been reduced mod p.

GaloisFields._conwaypolynomials

GaloisFields._factorization

GaloisFields._gcdx

GaloisFields._parse_declaration

GaloisFields._parsepoly

GaloisFields._rem

GaloisFields._upgrade

GaloisFields._widen

GaloisFields.basefield

GaloisFields.bitswap

GaloisFields.carrylessmul

GaloisFields.char — Function.

p = char(GaloisField(3)) # returns 3

Return the characteristic of a finite field, or 0 for <:Integer or <:Rational{<Integer}.

GaloisFields.conwaypolynomial

GaloisFields.cycliclogtable

GaloisFields.defaultshow — Function.

defaultshow(io, t)

Overloading display of types can be a bit hairy; I've seen a declaration like

show(..., ::Type{<:Val{B}}) where B

being called for non-concrete types. (I haven't dug deep enough to find a nice minimal example.)

That's why all show overloads for types do

!isconcretetype(t) && return defaultshow(io, t)

GaloisFields.disable_zech_multiplication

GaloisFields.enable_zech_multiplication

GaloisFields.eval

GaloisFields.expansion

GaloisFields.gen

GaloisFields.genname

GaloisFields.identifications

GaloisFields.identify

GaloisFields.include

GaloisFields.inttype

GaloisFields.isconway

GaloisFields.iter

GaloisFields.logtables

GaloisFields.minpoly

GaloisFields.minpolymask

GaloisFields.n

GaloisFields.overrides

GaloisFields.zech_op