Docstrings

StrTables.AbstractEntityTable — Type.

Abstract type for Entity tables: Supports lookupname, matchchar, matches, longestmatches, completions

StrTables.AbstractPackedTable

StrTables.BASE_CODE

StrTables.MAX_CODE

StrTables.PACKED_CODE

StrTables.PackedTable — Type.

Compact table Designed to save memory compared to a Vector{Vector{S}} Allows for fast lookup of ranges when input was sorted Can be saved/loaded to/from a file quickly

StrTables.STRING1_CODE

StrTables.STRING2_CODE

StrTables.STRING4_CODE

StrTables.STRTAB_CODE

StrTables.StrTable — Type.

Compact string table Designed to save memory compared to a Vector{String} Allows for fast lookup of ranges when input was sorted Can be saved/loaded to/from a file quickly

StrTables.StrTables — Module.

Simple String Table and load/save functions

Author: Scott Paul Jones Copyright: 2017-2018 Gandalf Software, Inc (and any future contributors) License: MIT (see https://github.com/JuliaString/StrTables.jl/blob/master/LICENSE.md)

Public API:

Exported:

StrTable takes a Vector of some AbstractString type and returns a compact string table, that acts as a AbstractVector{String}

Not exported:

matchfirst takes a sorted string table and a string, returns a vector of elements that start with the string

matchfirstrng takes a sorted string table and a string, returns the range of indexes of elements start with the string

save takes a filename and a collection (of simple types) and saves them to the file in a simple format

load takes a filename and returns a Vector of the values stored by save in the file

StrTables.SupTypes

StrTables.V6_COMPAT

StrTables.VECTOR_CODE

StrTables.VER

StrTables._codeunits

StrTables._contains

StrTables._empty_str

StrTables._empty_str_vec

StrTables._get_code

StrTables._get_str

StrTables._get_strings

StrTables._get_val2c

StrTables._getsize

StrTables._lexcmp

StrTables._ltvec

StrTables._memcmp

StrTables._replace

StrTables._veccmp

StrTables.completions — Function.

Given a string, return all of the entity names that start with that string, if any

StrTables.create_vector

StrTables.cvt_char

StrTables.eval

StrTables.include

StrTables.load — Function.

Return a Vector{Any} with data tables loaded from IO

Return a Vector{Any} with data tables loaded from file

StrTables.longestmatches — Function.

Given a string, return all of the longest matches to the beginning of the string as a vector

StrTables.lookupname — Function.

Given an entity name, return the string it represents, or an empty string if not found

StrTables.matchchar — Function.

Given a character, return all exact matches to the character as a vector

StrTables.matches — Function.

Given a string, return all exact matches to the string as a vector

StrTables.matchfirst — Function.

Return a vector of values that whose beginning matches the string

StrTables.matchfirstrng — Function.

Return the range of indices of values that whose beginning matches the string

StrTables.now

StrTables.pack_table

StrTables.parse_hex

StrTables.read_value — Function.

Read a single StrTable, String, UInt, or Vector of UInt value from the io device

StrTables.read_vector

StrTables.save — Function.

Save a collection of values (StrTable, String, Float, UInt, Int*) into an IO object

Save a collection of values (StrTable, String, Float, UInt, Int*) into a file

StrTables.sortsplit!

StrTables.type_tab

StrTables.write_value