Docstrings

Lasso.+

Lasso.-

Lasso.@formula

Lasso.AbstractContrasts

Lasso.AbstractMixtureModel

Lasso.AbstractMvNormal

Lasso.Arcsine

Lasso.Bernoulli

Lasso.Beta

Lasso.BetaBinomial

Lasso.BetaPrime

Lasso.Binomial

Lasso.Biweight

Lasso.CV1se

Lasso.CVmin

Lasso.Categorical

Lasso.CauchitLink

Lasso.Cauchy

Lasso.Chi

Lasso.Chisq

Lasso.CloglogLink

Lasso.CoefficientIterator

Lasso.Continuous

Lasso.ContinuousDistribution

Lasso.ContinuousMatrixDistribution

Lasso.ContinuousMultivariateDistribution

Lasso.ContinuousUnivariateDistribution

Lasso.ContrastsCoding

Lasso.CoordinateDescent

Lasso.Cosine

Lasso.CovarianceCoordinateDescent

Lasso.DensePred

Lasso.DensePredChol

Lasso.DensePredQR

Lasso.DiagNormal

Lasso.DiagNormalCanon

Lasso.Dirichlet

Lasso.DirichletMultinomial

Lasso.Discrete

Lasso.DiscreteDistribution

Lasso.DiscreteMatrixDistribution

Lasso.DiscreteMultivariateDistribution

Lasso.DiscreteUniform

Lasso.DiscreteUnivariateDistribution

Lasso.Distribution

Lasso.Distributions

Lasso.DoubleExponential

Lasso.DummyCoding

Lasso.EdgeworthMean

Lasso.EdgeworthSum

Lasso.EdgeworthZ

Lasso.EffectsCoding

Lasso.EmpiricalUnivariateDistribution

Lasso.Epanechnikov

Lasso.Erlang

Lasso.Estimator

Lasso.Exponential

Lasso.FDist

Lasso.FisherNoncentralHypergeometric

Lasso.Formula

Lasso.Frechet

Lasso.FullNormal

Lasso.FullNormalCanon

Lasso.FusedLasso

Lasso.FusedLassoMod

Lasso.GLM

Lasso.Gamma

Lasso.GammaLassoPath

Lasso.Gaussian

Lasso.GeneralizedExtremeValue

Lasso.GeneralizedLinearModel

Lasso.GeneralizedPareto

Lasso.Geometric

Lasso.GlmResp

Lasso.Gumbel

Lasso.HelmertCoding

Lasso.Hypergeometric

Lasso.IdentityLink

Lasso.InverseGamma

Lasso.InverseGaussian

Lasso.InverseLink

Lasso.InverseSquareLink

Lasso.InverseWishart

Lasso.IsoNormal

Lasso.IsoNormalCanon

Lasso.KSDist

Lasso.KSOneSided

Lasso.Kolmogorov

Lasso.Laplace

Lasso.Lasso

Lasso.LassoPath

Lasso.Levy

Lasso.LinPred

Lasso.LinPredModel

Lasso.LinearModel

Lasso.Link

Lasso.LmResp

Lasso.LocationScale

Lasso.LogLink

Lasso.LogNormal

Lasso.Logistic

Lasso.LogitLink

Lasso.MAX_DEV_FRAC

Lasso.MIN_DEV_FRAC_DIFF

Lasso.MLEstimator

Lasso.MatrixDistribution

Lasso.Matrixvariate

Lasso.MixtureModel

Lasso.ModelFrame

Lasso.ModelMatrix

Lasso.Multinomial

Lasso.Multivariate

Lasso.MultivariateDistribution

Lasso.MultivariateMixture

Lasso.MultivariateNormal

Lasso.MvLogNormal

Lasso.MvNormal

Lasso.MvNormalCanon

Lasso.MvNormalKnownCov

Lasso.MvTDist

Lasso.NaiveCoordinateDescent

Lasso.NegativeBinomial

Lasso.NegativeBinomialLink

Lasso.NonMatrixDistribution

Lasso.NoncentralBeta

Lasso.NoncentralChisq

Lasso.NoncentralF

Lasso.NoncentralHypergeometric

Lasso.NoncentralT

Lasso.Normal

Lasso.NormalCanon

Lasso.NormalInverseGaussian

Lasso.P

Lasso.Pareto

Lasso.Poisson

Lasso.PoissonBinomial

Lasso.ProbitLink

Lasso.QQPair

Lasso.RANDOMIZE_DEFAULT

Lasso.RandomCoefficientIterator

Lasso.Rayleigh

Lasso.RealInterval

Lasso.RegularizationPath

Lasso.S

Lasso.Sampleable

Lasso.Semicircle

Lasso.Skellam

Lasso.SparseCoefficients

Lasso.SqrtLink

Lasso.StatsModels

Lasso.SufficientStats

Lasso.SymTriangularDist

Lasso.TDist

Lasso.TrendFilter

Lasso.TrendFiltering

Lasso.TriangularDist

Lasso.Triweight

Lasso.Truncated

Lasso.TruncatedNormal

Lasso.Uniform

Lasso.Univariate

Lasso.UnivariateDistribution

Lasso.UnivariateGMM

Lasso.UnivariateMixture

Lasso.Util

Lasso.ValueSupport

Lasso.VariateForm

Lasso.VonMises

Lasso.VonMisesFisher

Lasso.WalleniusNoncentralHypergeometric

Lasso.Weibull

Lasso.Wishart

Lasso.ZeroMeanDiagNormal

Lasso.ZeroMeanDiagNormalCanon

Lasso.ZeroMeanFullNormal

Lasso.ZeroMeanFullNormalCanon

Lasso.ZeroMeanIsoNormal

Lasso.ZeroMeanIsoNormalCanon

Lasso.addcoef

Lasso.addcoef!

Lasso.addcoefs!

Lasso.adjr2

Lasso.adjr²

Lasso.aicc

Lasso.binaryentropy

Lasso.build_model

Lasso.canonform

Lasso.canonicallink

Lasso.ccdf

Lasso.cdf

Lasso.cdfit!

Lasso.cf

Lasso.cgf

Lasso.circvar

StatsBase.coef — Function.

coef(path::RegularizationPath; select=:AICc)

Returns a p by nλ coefficient array where p is the number of coefficients (including any intercept) and nλ is the number of path segments, or a selected segment's coefficients.

If model was only initialized but not fit, returns a p vector of zeros. Consistent with StatsBase.coef, if the model has an intercept it is included.

Example:

  m = fit(LassoPath,X,y)
  coef(m; select=:CVmin)

Keywords

select=:all returns a p by nλ matrix of coefficients

select=:AIC selects the AIC minimizing segment

select=:AICc selects the corrected AIC minimizing segment

select=:BIC selects the BIC minimizing segment

select=:CVmin selects the segment that minimizing out-of-sample mean squared error from cross-validation with nCVfolds random folds.

select=:CV1se selects the segment whose average OOS deviance is no more than 1 standard error away from the one minimizing out-of-sample mean squared error from cross-validation with nCVfolds random folds.

nCVfolds=10 number of cross-validation folds

kwargs... are passed to minAICc(::RegularizationPath) or to cross_validate_path(::RegularizationPath)

Lasso.coefnames

Lasso.coeftable

Lasso.component

Lasso.components

Lasso.componentwise_logpdf

Lasso.componentwise_pdf

Lasso.computeXssq

Lasso.computeXssqXty

Lasso.computeXtX!

Lasso.compute_grad

Lasso.compute_gradient

Lasso.computeyty

Lasso.computeλ

Lasso.computeμX!

Lasso.computeω!

Lasso.concentration

Lasso.confint

Lasso.copyω!

Lasso.cor

Lasso.cov

Lasso.cquantile

Lasso.cross_validate_path

Lasso.cumulant

Lasso.cycle!

Lasso.defaultalgorithm

Lasso.delbeta!

StatsBase.deviance — Function.

deviance at each segment of the path for the fitted model and data

deviance at each segement of the path for (potentially new) data X and y select=:all or :AICc like in coef()

deviance at each segment of the path for (potentially new) y and predicted values μ

Lasso.devresid

Lasso.dim

Lasso.dispersion_parameter

Lasso.distfun — Function.

distribution of underlying GLM

StatsBase.dof — Function.

dof(path::RegularizationPath)

Approximates the degrees-of-freedom in each segment of the path as the number of non zero coefficients plus a dispersion parameter when appropriate. Note that for GammaLassoPath this may be a crude approximation, as gamlr does this differently.

Lasso.dof_residual

Lasso.dropterm

Lasso.entropy

Lasso.estimate

Lasso.eval

Lasso.expected_logdet

Lasso.failprob

Lasso.fit

Lasso.fit!

Lasso.fit_map

Lasso.fit_map!

Lasso.fit_mle

Lasso.fit_mle!

Lasso.formula

Lasso.freecumulant

Lasso.ftest

Lasso.getXtX

Lasso.glm

Lasso.glmvar

Lasso.gradlogpdf

Lasso.hasfinitesupport

Lasso.hasintercept

Lasso.include

Lasso.infocrit

Lasso.initpenaltyfactor

Lasso.insupport

Lasso.intercept

Lasso.invcov

Lasso.inverselink

Lasso.invlogccdf

Lasso.invlogcdf

Lasso.invscale

Lasso.isbounded

Lasso.isleptokurtic

Lasso.islowerbounded

Lasso.ismesokurtic

Lasso.isplatykurtic

Lasso.isprobvec

Lasso.isupperbounded

Lasso.kde

Lasso.kurtosis

GLM.linkfun — Function.

link of underlying GLM

Lasso.linkinv

Lasso.linpred

Lasso.linpred!

Lasso.lm

Lasso.location

Lasso.location!

Lasso.logccdf

Lasso.logcdf

Lasso.logdetcov

Lasso.logistic

Lasso.logit

Lasso.loglikelihood

Lasso.logpdf

Lasso.logpdf!

Lasso.maxλ!

Lasso.mean

Lasso.meandir

Lasso.meanform

Lasso.meanlogx

Lasso.median

Lasso.mgf

Lasso.minAIC

Lasso.minAICc

Lasso.minBIC

Lasso.mode

Lasso.model_response

Lasso.modes

Lasso.moment

Lasso.mueta

Lasso.mustart

Lasso.ncategories

Lasso.ncomponents

Lasso.negbin

Lasso.nobs

Lasso.nsamples

Lasso.ntrials

Lasso.nullX

Lasso.nulldeviance

Lasso.nullloglikelihood

Lasso.params

Lasso.params!

Lasso.partype

Lasso.pathtype

Lasso.pdf

Lasso.poststep

Lasso.predict

Lasso.probs

Lasso.probval

Lasso.qqbuild

Lasso.quantile

Lasso.r2

Lasso.rate

Lasso.rescale

Lasso.residualoffset

Lasso.residuals

Lasso.r²

Lasso.sample

Lasso.sample!

Lasso.sampler

Lasso.scale

Lasso.scale!

Lasso.setcontrasts!

Lasso.shape

Lasso.skewness

Lasso.span

Lasso.sqmahal

Lasso.sqmahal!

Lasso.ssr

Lasso.standardizeX

Lasso.std

Lasso.stderror

Lasso.stdlogx

Lasso.succprob

Lasso.suffstats

Lasso.support

Lasso.test_distr

Lasso.test_samples

Lasso.update!

Lasso.update_coef!

Lasso.updateμ!

Lasso.var

Lasso.varlogx

Lasso.vcov

Lasso.weightedresiduals

Lasso.wrkresp

Lasso.wrkresp!

Lasso.wsample

Lasso.wsample!

Lasso.λω