Docstrings

NeighbourLists.AbstractIterator

NeighbourLists.CellList — Type.

CellList : store atoms in cells / bins. Mostly used internally to construct PairLists.

NeighbourLists.MAX_THREADS

NeighbourLists.NBodyIterator

NeighbourLists.NeighbourLists

NeighbourLists.PairIterator

NeighbourLists.PairList — Type.

PairList stores a neighbourlist as a list of pairs

PairList(X, cutoff, cell, pbc)
PairList(nlist::CellList)

where

X : positions, either as 3 x N matrix or as Vector{SVec}

cutoff : positive real value

cell : 3 x 3 matrix, with rows denoting the cell vectors

pbc : 3-tuple or array, storing periodicity information

Kw-args:

int_type : default is Int

store_first : whether to store the array of first indices, default true

sorted : whether to sort the j vector, default false

PairList fields

X, cutoff, i, j, r, R, first, where

(i[n], j[n]) denotes the indices of a neighbour pair, r[n] the distance between those atoms, R[n] the vectorial distance, note this is identical to X[i[n]]-X[j[n]] without periodic b.c.s, but with periodic boundary conditions it is different. first[m] contains the index to the first (i[n], j[n]) for which i[n] == first[m], i.e., (j, first) essentially defines a compressed column storage of the adjacancy matrix.

NeighbourLists.SMat

NeighbourLists.SVec

NeighbourLists.SiteIterator

NeighbourLists._celllist_

NeighbourLists._find_neighbours_!

NeighbourLists._find_next_ — Function.

function _find_next_(j, n, first)

j : array of neighbour indices

n : current site index

first : array of first indices

return the first index first[n] <= m < first[n+1] such that j[m] > n; and returns 0 if no such index exists

NeighbourLists._fix_cell_ — Function.

_fix_cell_(X::Vector{SVec{T}}, C::SMat{T}, pbc)

produces new X, C such that PBC are respected, but all positions are inside the cell. This Potentially involves

wrapping atom positions in the pbc directions

shifting atom positions in the non-pbc directions

enlarging the cell

If either X or C are modified, then they will be copied.

NeighbourLists._item

NeighbourLists._mt_map_!

NeighbourLists._pairlist_

NeighbourLists._sub2ind

NeighbourLists.analyze_cell

NeighbourLists.bin_trunc — Function.

Map i back to the interval [0,n) by assigning edge value if outside interval

apply bin_trunc only if open bdry

NeighbourLists.bin_wrap — Function.

Map i back to the interval [0,n) by shifting by integer multiples of n

apply bin_wrap only if periodic bdry

NeighbourLists.bin_wrap_or_trunc — Function.

applies bintrunc to open bdry and binwrap to periodic bdry

NeighbourLists.eval

NeighbourLists.get_first — Function.

get_first(i::Vector{TI}, nat::Integer) -> first::Vector{TI} where TI <: Integer

Assumes that i is sorted in ascending order. For n = 1, . . ., nat, first[n] will be the index to the first element of i with value n. Further, first[nat+1] will be length(i) + 1.

If first[n] == first[n+1] then this means that i contains no element n.

NeighbourLists.inc

NeighbourLists.include

NeighbourLists.lengths

NeighbourLists.maptosites!

NeighbourLists.maptosites_d!

NeighbourLists.maptosites_d_inner! — Function.

mapreduce_antisym!{T}(out::AbstractVector{SVec{T}}, df, it::PairIterator{T})

anti-symmetric variant of mapreduce!{S, T}(out::AbstractVector{S}, ...), summing only over bonds (i,j) with i < j and adding f(Rij) to site j and -f(Rij) to site i.

NeighbourLists.maptosites_inner! — Function.

mapreduce_sym!{S, T}(out::AbstractVector{S}, f, it::PairIterator{T})

symmetric variant of mapreduce!{S, T}(out::AbstractVector{S}, ...), summing only over bonds (i,j) with i < j and adding f(R_ij) to both sites i, j.

NeighbourLists.max_neigs — Function.

max_neigs(nlist::PairList) -> Integer

returns the maximum number of neighbours that any atom in the neighbourlist has.

NeighbourLists.mt_split

NeighbourLists.mt_split_interlaced

NeighbourLists.nbodies — Function.

nbodies(N, nlist::PairList)

creates an N-body iterator, e.g., nbodies(3, nlist) will create an iterator over permutation-symmetric 3-body terms. Use mapreduce_sym! and map_reduce_sym_d! to carry out the iterations.

NeighbourLists.npairs

NeighbourLists.nsites

NeighbourLists.pairs

NeighbourLists.position_to_cell_index — Function.

Map particle position to a (cartesian) cell index

NeighbourLists.set_maxthreads!

NeighbourLists.setup_mt

NeighbourLists.site

NeighbourLists.sites

NeighbourLists.sort_neigs! — Function.

sort_neigs!(j, r, R, first)

sorts each sub-range of j corresponding to one site in ascending order and applies the same permutation to r, R, S.