Docstrings

VisualRegressionTests.@plottest

VisualRegressionTests.@visualtest

VisualRegressionTests.CLOSE_MATCH

VisualRegressionTests.DOES_NOT_MATCH

VisualRegressionTests.EXACT_MATCH

VisualRegressionTests.PROCESSING_ERROR

VisualRegressionTests.VisualRegressionTests

VisualRegressionTests.VisualTest

VisualRegressionTests.VisualTestResult

VisualRegressionTests.VisualTestStatus

VisualRegressionTests.__init__

VisualRegressionTests.accum

VisualRegressionTests.blurdiff — Function.

blur images A and B with sigma and then compute a difference

VisualRegressionTests.compare_images

VisualRegressionTests.difftype

VisualRegressionTests.eval

VisualRegressionTests.include

VisualRegressionTests.maxabsfinite — Function.

maximum absolute value in A ignoring Inf or NaN

VisualRegressionTests.sad

VisualRegressionTests.sumdiff

VisualRegressionTests.test_images