Docstrings

SimpleANOVA.AnovaData — Type.

AnovaData

Container for the complete results of an ANOVA test.

SimpleANOVA.AnovaEffect — Type.

AnovaEffect

Supertype for all ANOVA data items: AnovaValue, AnovaFactor, AnovaResult

SimpleANOVA.AnovaFactor — Type.

AnovaFactor <: AnovaEffect

A set of values for an Anova effect which has a mean square.

name - the name of this factor

ss - sum of squares

df - degrees of freedom

ms - mean square

SimpleANOVA.AnovaResult — Type.

AnovaResult <: AnovaEffect

A set of values for an Anova factor which has been tested

name - the name of this factor

ss - sum of squares

df - degrees of freedom

ms - mean square

f - the F statistic

p - the probability of a Type I error (imcorrect rejection of null hypothesis)

SimpleANOVA.AnovaValue — Type.

AnovaValue <: AnovaEffect

A set of values for an Anova item for which a mean square is not required.

name - the name of this value

ss - sum of squares

df - degrees of freedom

SimpleANOVA.FactorType — Type.

FactorType

fixed - A crossed fixed-effects factor

random - A crossed random-effects factor

nested - A random factor fully nested within another factor

source

SimpleANOVA.InvertedIndex — Type.

InvertedIndex(idx)
Not(idx)

Construct an inverted index, selecting all indices not in the passed idx.

Upon indexing into an array, the InvertedIndex behaves like a 1-dimensional collection of the indices of the array that are not in idx. Bounds are checked to ensure that all indices in idx are within the bounds of the array — even though they are skipped. The InvertedIndex behaves like a 1-dimensional collection of its inverted indices. If idx spans multiple dimensions (like a multidimensional logical mask or CartesianIndex), then the inverted index will similarly span multiple dimensions.

Copied and trimmed down from InvertedIndices.jl repository

SimpleANOVA.Not

SimpleANOVA.SimpleANOVA

SimpleANOVA.__init__

SimpleANOVA.amongnestedfactorscalc

SimpleANOVA.anova — Function.

anova(observations::Array{Union{Number, Vector{Number}}}, factortypes = FactorType[]; factornames = String[], hasreplicates = true)
anova(observations::Vector{Number}, factorassignments::Vector{Vector{Any}}, factortypes = FactorType[]; factornames = String[], hasreplicates = true)
anova(df::DataFrame, observationscolumn::Symbol, factorcolumns::Vector{Symbol}, factortypes = FactorType[]; factornames = String[])

Performs an Analysis of Variance (ANOVA) calculation.

Operates on up to 3 crossed factors (fixed or random) and arbitrarily many random nested factors, with or without replicates, on balanced data.

Arguments

observations: Array containing the values to test. For the array, each dimension is a factor level, such that observations[2,5,3] indicates the 2nd level of the first factor, the 5th level of the second factor, and the 3rd level of the third factor. May contain values or vectors of values, where the vector contains replicates. Factors should be ordered with least significant first. For the vector, must provide factorassignments to specify factor levels.

factorassignments: Vector of vectors of integers specifying how each observation is assigned to a factor level. Provide this when observations is given as a vector. Factor levels do not have to be consecutive or ordered. Nested factors must reuse factor levels currently.

factortypes: Vector indicating the FactorType for each factor. If present, replicates must appear first, any nested after, and then random or fixed in any order. Specify replicates if the first dimension of the observations matrix contains replicate values (vs. contained in vectors). If too few values are provided, remaining are assumed to be fixed.

factornames: Vector of names for each factor, excluding the replicate factor. If empty, will be automatically populated alphabetically.

Notes: The last index will be the top factor in the table.

Output: AnovaData structure containing the test results for each factor.

Examples

anova(observations)                        # N-way fixed-effects ANOVA with replicates (vectors or first dimension)
anova(observations, hasreplicates = false) # N-way fixed-effects ANOVA without replicates (first dimension)
anova(observations, [random])              # N-way ANOVA with lower random factor and 1 or 2 upper fixed factors
anova(observations, [random])              # N-way ANOVA with lower random factor and 1 or 2 upper fixed factors
anova(observations, [fixed, random])       # N-way ANOVA with 1 lower fixed factor, 1 random factor, and 0 or 1 upper fixed factor
anova(observations, [nested, random])      # N-way fixed-effects ANOVA with 1 random nested factor, 1 random factor, and 1-2 fixed factors

Glossary

observation: The dependent variable.

factor: An independent variable.

factor level: A value of a factor.

balanced: All combinations of factor levels have the same number of observations.

crossed factor: A factor with levels that combine with the levels of all other crossed factors.

fixed factor: A factor with fixed effects (e.g. treatment, concentration, exposure time).

random factor: A factor with random effects (e.g. location, individual).

nested factor: A random factor where the levels are unique to a combination of crossed factor levels (e.g. replicate).

sum of squares (SS): A measure of variance that is dependent on sample size. Also called "sum of squared deviations."

degrees of freedom (DF, ν): The number of bins in which the values could have been moved, if random.

mean square (MS): SS / DF. Corrects for the larger variance expected if random values can be assigned to more bins. Also called "mean squared error" or "mean squared deviation."

F-statistic: The division of MS values produce a result belonging to the "F distribution", the shape of which depends on the DF of the numerator and denominator. The location of this value on the distribution provides the p-value.

p-value: The probability that, if all measurements had been drawn from the same population, you would obtain data at least as extreme as contained in your observations.

SimpleANOVA.anovakernel

SimpleANOVA.cellscalc

SimpleANOVA.cellsname

SimpleANOVA.errorcalc

SimpleANOVA.errorname

SimpleANOVA.eval

SimpleANOVA.factorscalc

SimpleANOVA.fixed

SimpleANOVA.ftest

SimpleANOVA.getdenominators

SimpleANOVA.getnumerators

SimpleANOVA.include

SimpleANOVA.interactionscalc

SimpleANOVA.nested

SimpleANOVA.nestedfactorscalc

SimpleANOVA.pairwisecalc

SimpleANOVA.plot

SimpleANOVA.random

SimpleANOVA.remaindername

SimpleANOVA.spanned_indices

SimpleANOVA.sumfirstdim

SimpleANOVA.threewayinteraction

SimpleANOVA.threewisecalc

SimpleANOVA.totalcalc

SimpleANOVA.totalname

SimpleANOVA.validate