Docstrings

StructIO.@io — Macro.

@io <type definition>
    ...
end

Generates packing_strategy() and packed_sizeof() methods for the type being defined within the given type definition. This enables usage of the unpack method.

StructIO.Default

StructIO.Packed

StructIO.PackingStrategy

StructIO.StructIO

StructIO.bswap! — Function.

bswap!(ptr::Ptr{UInt8}, sz)

Byte-swap a chunk of data in-place

StructIO.eval

StructIO.fieldsize — Function.

fieldsize(T::DataType, field_idx)

Return the size (in bytes) of a field within T in memory

StructIO.fix_endian — Function.

fix_endian(x, endianness::Symbol)

Returns a byte-swapped version of x if the given endianness must be swapped for the current host system.

StructIO.include

StructIO.needs_bswap — Function.

needs_bswap(endianness::Symbol)

Returns true if the given endianness does not match the current host system.

StructIO.pack — Function.

pack(io::IO, source, endianness::Symbol = :NativeEndian)

Given an input source, pack it into io, byte-swapping according to the given endianness of io. If endianness is :NativeEndian (the default), no byteswapping will occur. If endianness is :LittleEndian or :BigEndian, byteswapping will occur if the endianness of the host system does not match the endianness of io.

StructIO.packed_sizeof

StructIO.packing_strategy — Function.

packing_strategy(x)

Return the packing strategy for the given type, defaults to Default, is overridden by auto-generated methods for specific types from @io invocations.

StructIO.unpack — Function.

unpack(io::IO, T::Type, endianness::Symbol = :NativeEndian)

Given an input io, unpack type T, byte-swapping according to the given endianness of io. If endianness is :NativeEndian (the default), no byteswapping will occur. If endianness is :LittleEndian or :BigEndian, byteswapping will occur of the endianness if the host system does not match the endianness of io.

StructIO.unsafe_pack — Function.

unsafe_pack(io, source, endianness, ::Type{Packed/Default})

Pack source into io, byte-swapping if endianness dictates we should. The last argument is a packing strategy, used to determine the layout of the data in memory. All Packed objects recurse until bitstypes objects are eventually reached, at which point Default packing is identical to Packed behavior.

StructIO.unsafe_unpack — Function.

unsafe_unpack(io, T, target, endianness, ::Type{Default})

Unpack an object of type T from io into target, byte-swapping if endianness dictates we should, assuming a Default packing strategy. All packed structs recurse until bitstypes objects are eventually reached, at which point Default packing is the only behavior.