Docstrings

GaussianMixtures.API

GaussianMixtures.CSstats — Type.

CSstats a type holding centered and scaled zeroth and first order GMM statistics

GaussianMixtures.CovType

GaussianMixtures.Cstats — Type.

Cstats, a type holding centered zeroth, first and second order GMM statistics

GaussianMixtures.Data — Type.

Data is a type for holding an array of feature vectors (i.e., matrices), or references to files on disk. The data is automatically loaded when needed, e.g., by indexing.

GaussianMixtures.DataOrMatrix

GaussianMixtures.DiagCov

GaussianMixtures.FullCov

GaussianMixtures.GMM — Type.

GMM is the type that stores information of a Guassian Mixture Model. Currently two main covariance types are supported: full covarariance and diagonal covariance.

GaussianMixtures.GMM2

GaussianMixtures.GMMk

GaussianMixtures.GMMprior — Type.

GMMprior is a type that holds the prior for training GMMs using Variational Bayes.

GaussianMixtures.GaussianMixture — Type.

GaussianMixture, an abstract type for a mixture of full-covariance or diagonal-covariance Gaussian distributions

GaussianMixtures.GaussianMixtures

GaussianMixtures.History — Type.

History, a type to record the history of how a GMM is built.

GaussianMixtures.MatOrVecMat

GaussianMixtures.UTriangular

GaussianMixtures.VGMM — Type.

VGMM is the type that is used to store a GMM in the Variational Bayes training.

GaussianMixtures.VecOrMat

GaussianMixtures._load

GaussianMixtures._size

GaussianMixtures.addhist! — Function.

addhist!(::GMM, s) adds a comment s to the GMMM

GaussianMixtures.avll

GaussianMixtures.binsearch

GaussianMixtures.cholinv

GaussianMixtures.compute_range

GaussianMixtures.covar — Function.

covar(GMM.Σ) extracts the covariances Σ (which may be encoded as chol(inv(Σ))

GaussianMixtures.covars — Function.

covars(::GMM) returns the covariance matrices Σ of the Gaussians in the mixture.

GaussianMixtures.csstats

GaussianMixtures.cstats

GaussianMixtures.density — Function.

density(gmm::GMM, X) returns log(P(X|μ, Σ))

GaussianMixtures.dmap

GaussianMixtures.dmapreduce

GaussianMixtures.dotscore

GaussianMixtures.em!

GaussianMixtures.emstep!

GaussianMixtures.eval

GaussianMixtures.expectations

GaussianMixtures.eye

GaussianMixtures.forcesymmetric

GaussianMixtures.full — Function.

full(::GMM) turns a diagonal covariance GMM into a full-covariance GMMM

GaussianMixtures.gmmposterior

GaussianMixtures.gmmsplit

GaussianMixtures.history

GaussianMixtures.include

GaussianMixtures.kind — Function.

kind(::GMM) returns the kind of GMM, either :diag or :full

GaussianMixtures.llpg

GaussianMixtures.logsumexp

GaussianMixtures.logsumexpw

GaussianMixtures.logρ

GaussianMixtures.lowerbound

GaussianMixtures.maxapost

GaussianMixtures.means — Function.

means(::GMM) returns the means μ of the Gaussians in the mixture

GaussianMixtures.mem

GaussianMixtures.mstep

GaussianMixtures.mylogdet

GaussianMixtures.n_components

GaussianMixtures.nparams — Function.

nparams(::GMM) returns the number of free parameters in the GMM

GaussianMixtures.precision

GaussianMixtures.rand

GaussianMixtures.retranspose

GaussianMixtures.rmdisfunct

GaussianMixtures.rnk

GaussianMixtures.sanitycheck!

GaussianMixtures.setmem

GaussianMixtures.show

GaussianMixtures.stats

GaussianMixtures.trAB

GaussianMixtures.vec

GaussianMixtures.weights — Function.

weights(::GMM) returns the weights w, or priors, of the Gaussians in the mixture

GaussianMixtures.xμTΛxμ!