Docstrings

MeshCat.@pack_AnimationClip

MeshCat.@pack_AnimationClip!

MeshCat.@pack_GenericMaterial

MeshCat.@pack_GenericMaterial!

MeshCat.@pack_PointsMaterial

MeshCat.@pack_PointsMaterial!

MeshCat.@pack_Texture

MeshCat.@pack_Texture!

MeshCat.@unpack_AnimationClip

MeshCat.@unpack_GenericMaterial

MeshCat.@unpack_PointsMaterial

MeshCat.@unpack_Texture

MeshCat.ASSET_KEYS

MeshCat.AbstractCommand

MeshCat.AbstractControl

MeshCat.AbstractMaterial

MeshCat.AbstractObject

MeshCat.AbstractVisualizer

MeshCat.Animation

MeshCat.AnimationClip

MeshCat.AnimationFrameVisualizer

MeshCat.AnimationTrack

MeshCat.Button

MeshCat.CoreVisualizer

MeshCat.Delete

MeshCat.GenericMaterial

MeshCat.GeometryLike

MeshCat.HyperCylinder

MeshCat.HyperEllipsoid

MeshCat.IJuliaCell

MeshCat.LineBasicMaterial

MeshCat.LineSegments

MeshCat.Mesh

MeshCat.MeshBasicMaterial

MeshCat.MeshCat

MeshCat.MeshLambertMaterial

MeshCat.MeshPhongMaterial

MeshCat.NumericControl

MeshCat.Object

MeshCat.PackedVector

MeshCat.PngImage

MeshCat.PointCloud

MeshCat.Points

MeshCat.PointsMaterial

MeshCat.SceneTrees

MeshCat.SetAnimation

MeshCat.SetControl

MeshCat.SetObject

MeshCat.SetProperty

MeshCat.SetTransform

MeshCat.Texture

MeshCat.Triad

MeshCat.VIEWER_ROOT

MeshCat.ViewerWindow

MeshCat.Visualizer — Type.

vis = Visualizer()

Construct a new MeshCat visualizer instance.

Useful methods:

vis[:group1] # get a new visualizer representing a sub-tree of the scene
setobject!(vis, geometry) # set the object shown by this visualizer's sub-tree of the scene
settransform!(vis], tform) # set the transformation of this visualizer's sub-tree of the scene

MeshCat.__init__

MeshCat._setprop!

MeshCat.atframe

MeshCat.center

MeshCat.convert_frames_to_video

Base.delete! — Function.

delete!(vis)

Delete the geometry at this visualizer's path and all of its descendants.

MeshCat.develop_meshcat_assets

MeshCat.eval

MeshCat.extcode

MeshCat.flatten!

MeshCat.geometry

MeshCat.getclip!

MeshCat.include

MeshCat.intrinsic_transform — Function.

intrinsic_transform(g)

Different tools disagree about what various geometric primitives mean. For example, GeometryTypes.jl considers the "origin" of a cube to be its bottom-left corner, where DrakeVisualizer and MeshCat consider its origin to be the center. The intrinsic_transform(g) returns the transform from the GeometryTypes origin to the MeshCat origin.

MeshCat.js_array_type

MeshCat.js_position

MeshCat.js_quaternion

MeshCat.lower — Function.

Convert a geometry, material, object, or transform into the appropriate plain data structures expected by three.js. Most objects are lowered into Dicts matching the JSON structure used by three.js.

MeshCat.material

MeshCat.open_url

MeshCat.radii

MeshCat.replace_with_uuid!

MeshCat.send

MeshCat.send_scene

MeshCat.setanimation!

MeshCat.setcontrol!

MeshCat.setobject! — Function.

setobject!(vis, obj)

Set the object at this visualizer's path. This replaces whatever geometry was presently at that path. To draw multiple geometries, place them at different paths by using the slicing notation:

setobject!(vis[:group1][:box1], geometry1)
setobject!(vis[:group1][:box2], geometry2)

MeshCat.setprop! — Function.

setprop!(vis, property, value)

Set a single property for the object at the given path.

(this is named setprop! instead of setproperty! to avoid confusion with the Base.setproperty! function introduced in Julia v0.7)

MeshCat.settransform! — Function.

settransform!(vis, tform)

Set the transform of this visualizer's path. This can be done before or after adding an object at that path. The overall transform of an object is the composition of the transforms of all of its parents, so setting the transform of vis[:group1] affects the poses of the objects at vis[:group1][:box1] and vis[:group1][:box2].

MeshCat.threejs_type

MeshCat.to_zero_index

MeshCat.update_tree!

MeshCat.wider_js_type