Docstrings

StataDTAFiles.ByteOrder

StataDTAFiles.ByteOrderIO

StataDTAFiles.DATANUMTYPES — Constant.

Number types is Stata that correspond to native Julia types, and are denoted by the latter.

StataDTAFiles.DTAFile — Type.

struct DTAFile{T<:NamedTuple, B<:StataDTAFiles.ByteOrderIO}

Representation of a Stata DTA file for which everything except the data itself has been read.

The data (rows) can be read as NamedTuples, using the iteration interface.

StataDTAFiles.DTAHeader — Type.

struct DTAHeader

DTA file header (without the byte order, which is encoded in the corresponding ByteOrderIO.

StataDTAFiles.DTAMap

StataDTAFiles.EPOCH — Constant.

Start of the epoch, ie "day 0" for most date handling functions in Stata.

Note

Don't use directly, see elapsed_days.

StataDTAFiles.EXTRANUMTYPES — Constant.

Number types which are read by readnum, in addition to DATANUMTYPES.

StataDTAFiles.LSF

StataDTAFiles.MAXFLOAT32

StataDTAFiles.MAXFLOAT64

StataDTAFiles.MAXINT16

StataDTAFiles.MAXINT32

StataDTAFiles.MAXINT8

StataDTAFiles.MSF

StataDTAFiles.READNUMTYPES — Constant.

All number types read by readnum.

StataDTAFiles.STRFSMAXLEN — Constant.

Maximum length of str# (aka strfs) strings in Stata DTA files.

StataDTAFiles.StataDTAFiles

StataDTAFiles.TIMESTAMPFMT — Constant.

Date format of Stata file timestamps.

StataDTAFiles._readrow

StataDTAFiles.decode_missing

StataDTAFiles.decode_variable_type — Function.

decode_variable_type(code)

Map the numerical type code in a DTA file to a Julia type.

Note that numeric types use a sentinel for missing values, and thus are decoded as Union{Missing, T} for the corresponding type T.

StataDTAFiles.elapsed_days — Function.

elapsed_days(Δ)

Convert a Stata "elapsed date" representation into Date.

Corresponds to the %td format in Stata.

StataDTAFiles.eval

StataDTAFiles.include

StataDTAFiles.read_formats — Function.

read_formats(boio, header, map)

Read the format strings, returning a Vector{String}.

StataDTAFiles.read_header

StataDTAFiles.read_map

StataDTAFiles.read_sortlist — Function.

read_sortlist(boio, header, map)

Read the sortlist, a vector if integers which contains column indexes for nested sorting.

StataDTAFiles.read_variable_names — Function.

read_variable_names(boio, header, map)

Read variable names, returning a tuple of Symbols.

StataDTAFiles.read_variable_types — Function.

read_variable_types(boio, header, map)

Read variable types from a DTA file. Return a tuple of Julia types.

StataDTAFiles.readbyteorder

StataDTAFiles.readchompedstring — Function.

readchompedstring(boio, len)

Read chars into a buffer of length len, find the terminating " " (if any) and truncate, returning a string.

readchompedstring(boio, T)

Read length (of type T), then read and chomp the string.

StataDTAFiles.readfield

StataDTAFiles.readfixedstring

StataDTAFiles.readnum

StataDTAFiles.readrow

StataDTAFiles.verifytag