Docstrings

IntervalOptimisation.IntervalOptimisation

IntervalOptimisation.SortedVectors

IntervalOptimisation.eval

IntervalOptimisation.include

IntervalOptimisation.maximise

IntervalOptimisation.maximize

IntervalOptimisation.minimise — Function.

minimise(f, X, tol=1e-3)

Find the global minimum of the function f over the Interval or IntervalBox X using the Moore-Skelboe algorithm.

For higher-dimensional functions $f:\mathbb{R}^n \to \mathbb{R}$, f must take a single vector argument.

Returns an interval containing the global minimum, and a list of boxes that contain the minimisers.

IntervalOptimisation.minimize — Function.

minimise(f, X, tol=1e-3)

Find the global minimum of the function f over the Interval or IntervalBox X using the Moore-Skelboe algorithm.

For higher-dimensional functions $f:\mathbb{R}^n \to \mathbb{R}$, f must take a single vector argument.

Returns an interval containing the global minimum, and a list of boxes that contain the minimisers.