Docstrings

StrBase.ASCIIStr

StrBase.ASCIISubStr

StrBase.AccessType

StrBase.BinaryStr

StrBase.CHUNKMSK

StrBase.CHUNKSZ

StrBase.CT

StrBase.CanContain — Function.

CanContain(Union{A, typeof(A)}, Union{B, typeof(B)})

CanContainStyle specifies whether the first string can contain a substring of the second type, and if so, what is the most efficient method of comparison

CanContain(::Type{<:MyString}, ::Type{String}) = ByteCompare()

Returns an instance of type CompareStyle, default CodePointCompare

StrBase.CharStat

StrBase.CharTypes

StrBase.EqualsStyle — Function.

EqualsStyle(Union{A, typeof(A)}, Union{B, typeof(B)})

EqualsStyle specifies whether the first string can equal a substring of the second type, and if so, what is the most efficient method of comparison

This is determined by the CanContain trait Returns an instance of type CompareStyle, default CodePointCompare

StrBase.HAS_WMEM

StrBase.ImmutableStr

StrBase.LatinStr

StrBase.LatinSubStr

StrBase.LineCounts

StrBase.MCU

StrBase.MS_ASCIILatin

StrBase.MS_ByteStr

StrBase.MS_Latin

StrBase.MS_RawUTF8

StrBase.MS_SubUTF32

StrBase.MS_UTF16

StrBase.MS_UTF32

StrBase.MS_UTF8

StrBase.MultiCU

StrBase.MutableStr

StrBase.MutableStyle — Type.

MutableStyle(A)
MutableStyle(typeof(A))

MutableStyle specifies the whether a string type is mutable or not

MutableStyle(::Type{<:MyString}) = MutableStr()

The default is ImmutableStr

StrBase.OthChr

StrBase.RawUTF16Str

StrBase.RawUTF8Str

StrBase.SCU

StrBase.STR_KEEP_NUL

StrBase.SetOfChars

StrBase.SingleCU

StrBase.Str

StrBase.StrBase — Module.

StrBase package

Copyright 2017-2018 Gandalf Software, Inc., Scott P. Jones, and other contributors to the Julia language Licensed under MIT License, see LICENSE.md Based partly on code in LegacyStrings that used to be part of Julia

StrBase.Text1Str

StrBase.Text2Str

StrBase.Text4Str

StrBase.UCS2Str

StrBase.UInt16_S

StrBase.UInt16_U

StrBase.UInt16_US

StrBase.UInt32_S

StrBase.UInt32_U

StrBase.UInt32_US

StrBase.UTF16Str

StrBase.UTF32Str

StrBase.UTF8Str

StrBase.UTF_INVALID

StrBase.UTF_LATIN1

StrBase.UTF_LONG

StrBase.UTF_SURROGATE

StrBase.UTF_UNICODE2

StrBase.UTF_UNICODE3

StrBase.UTF_UNICODE4

StrBase.UniRawChar

StrBase.UniStr — Constant.

Union type for fast dispatching

StrBase.WidChr

StrBase.WideCodeUnit

StrBase._LatinStr

StrBase._LatinSubStr

StrBase._UCS2Str

StrBase._UTF32Str

StrBase.__api__

StrBase.__replace

StrBase.__rsplit

StrBase.__split

StrBase.__tmp_api__

StrBase._align_len_utf16

StrBase._align_len_utf8

StrBase._aligned_set

StrBase._all_bmp

StrBase._all_latin

StrBase._allocate

StrBase._ascii_mask

StrBase._bmp_mask

StrBase._bmp_mask_32

StrBase._calc_len

StrBase._calcpnt

StrBase._check_bloom_mask

StrBase._check_bmp_utf16_al

StrBase._check_bmp_utf16_ul

StrBase._check_bmp_utf8_al

StrBase._check_bmp_utf8_ul

StrBase._check_char

StrBase._check_latin_utf8_al

StrBase._check_latin_utf8_ul

StrBase._check_mask_al

StrBase._check_mask_ul

StrBase._check_uppercase

StrBase._check_utf8

StrBase._check_utf8_al

StrBase._check_utf8_rest

StrBase._cmp

StrBase._cmp_str — Function.

Compare two strings, starting at nxtstr and nxtsub

StrBase._cmp_utf16

StrBase._cnt_non_bmp

StrBase._collectstr

StrBase._convert

StrBase._convert_ascii

StrBase._convert_utf_n

StrBase._copysub

StrBase._count_cont

StrBase._count_mask_al

StrBase._count_mask_ul

StrBase._cpcmp

StrBase._cpeq

StrBase._cpyconvert

StrBase._cpysubset

StrBase._cvt_16_to_utf8

StrBase._cvt_utf8

StrBase._cvtsize

StrBase._data — Function.

Codeunits of string as a Vector

StrBase._empty_sub

StrBase._encode_ascii_latin

StrBase._encode_char_utf8

StrBase._encode_utf16 — Function.

Converts an already validated UTF-8 encoded vector of UInt8 to a UTF16Str

Input Arguments:

pnt Ptr{UInt8} of UTF-8 encoded data

len length of output in 16-bit words

Returns:

UTF16Str

Converts an already validated UTF-32 encoded vector of UInt32 to a UTF16Str

Input Arguments:

dat Vector{UInt32} of UTF-32 encoded data

len length of output in 16-bit words

Returns:

Vector{UInt8}

StrBase._encode_utf32

StrBase._encode_utf8 — Function.

Converts an already validated vector of UInt16 or UInt32 to a UTF8Str

Input Arguments:

dat Vector of code units (UInt16 or UInt32)

len length of output in bytes

Returns:

UTF8Str

StrBase._fasteq

StrBase._fastwrite

StrBase._fwd_memchr

StrBase._get_bmp_mask

StrBase._get_dir

StrBase._get_lead

StrBase._get_masked

StrBase._getindex

StrBase._hash

StrBase._hash_abs

StrBase._hi_bit_16

StrBase._index

StrBase._is_latin_ucs2

StrBase._iseq

StrBase._ismutable

StrBase._isvalid_char_pos

StrBase._iterate

StrBase._iterate_utf8

StrBase._join

StrBase._joincvt

StrBase._joinio

StrBase._lastindex

StrBase._latin_mask

StrBase._latin_to_utf8

StrBase._length

StrBase._length_al

StrBase._length_check

StrBase._lenpntfin

StrBase._load

StrBase._lower

StrBase._lower_utf8

StrBase._lpad

StrBase._map

StrBase._maprest — Function.

Handle case where result vector is longer

StrBase._mask_allsurr

StrBase._mask_bytes

StrBase._mask_surr

StrBase._memcmp

StrBase._memcmp16

StrBase._memcpy

StrBase._memhash

StrBase._memset

StrBase._msk16

StrBase._msk32

StrBase._mskdn16

StrBase._mskdn32

StrBase._mskup16

StrBase._mskup32

StrBase._narrow!

StrBase._next

StrBase._nextcp

StrBase._nextcpfun

StrBase._nextind

StrBase._normalize

StrBase._not_found

StrBase._occurs_in

StrBase._pnt64

StrBase._prevind

StrBase._repeat

StrBase._repeat_3

StrBase._repeat_chr

StrBase._repeat_str

StrBase._ret_check

StrBase._rev_memchr

StrBase._reverse

StrBase._rpad

StrBase._rs

StrBase._search_bloom_mask

StrBase._srch_codeunit

StrBase._srch_cp

StrBase._srch_pred

StrBase._srch_str_bloom

StrBase._srch_strings

StrBase._str

StrBase._str_cpy

StrBase._string

StrBase._thisind

StrBase._trail_mask

StrBase._transcode_utf16_to_utf32

StrBase._transcode_utf8

StrBase._transcode_utf8_to_utf32

StrBase._ul

StrBase._unsafe_str8

StrBase._upper

StrBase._upper_utf8

StrBase._utf8_to_latin

StrBase._wide_lower_latin

StrBase._wide_upper

StrBase._widen!

StrBase._wideneq

StrBase._widenupper

StrBase._wrap_substr

StrBase.adjust_utf16

StrBase.alignedtype

StrBase.alignpnt

StrBase.calc_type

StrBase.calcstats

StrBase.check_3byte

StrBase.check_continuation

StrBase.check_string — Function.

Validates and calculates number of characters in a UTF-8,UTF-16 or UTF-32 encoded vector/string

This function checks the bounds of the start and end positions Use unsafe_check_string to avoid that overhead if the bounds have already been checked

Input Arguments:

dat UTF-8 (Vector{UInt8}), UTF-16 (Vector{UInt16}) or UTF-32 (Vector{UInt32}, AbstractString) encoded string

Optional Input Arguments:

startpos start position (defaults to 1)

endpos end position (defaults to lastindex(dat))

Keyword Arguments:

accept_long_null = false # Modified UTF-8 (\0 represented as b"\xc0\x80")

accept_surrogates = false # CESU-8

accept_long_char = false # Accept arbitrary long encodings

accept_invalids = false # Accept invalid sequences (to be replaced on conversion)

Returns:

(total characters, flags, 4-byte, 3-byte, 2-byte)

Throws:

StringError

StrBase.check_utf8 — Function.

Return index of first invalid codeunit (negative), 0 if all ASCII, or index of first non-ASCII codeunit

StrBase.check_valid

StrBase.checkcont

StrBase.checkkeep

StrBase.count_chars — Function.

Calculate the total number of characters, as well as number of latin1, 2-byte, 3-byte, and 4-byte sequences in a validated UTF-8 string

Calculate the total number of characters, as well as number of latin1, 2-byte, 3-byte, and 4-byte sequences in a validated UTF-16, UCS2, or UTF-32 string

StrBase.count_latin — Function.

Calculate the total number of bytes > 0x7f

StrBase.countmask

StrBase.cvthash

StrBase.empty_ascii

StrBase.empty_binary

StrBase.empty_latin

StrBase.empty_rawutf16

StrBase.empty_rawutf8

StrBase.empty_str

StrBase.empty_string

StrBase.empty_text1

StrBase.empty_text2

StrBase.empty_text4

StrBase.empty_ucs2

StrBase.empty_utf16

StrBase.empty_utf32

StrBase.empty_utf8

StrBase.endparse

StrBase.eq_bytes

StrBase.eval

StrBase.fast_check_string

StrBase.get_ch

StrBase.get_codeunit

StrBase.get_cp

StrBase.get_supplementary

StrBase.get_utf8_2byte

StrBase.get_utf8_3byte

StrBase.get_utf8_4byte

StrBase.hi_mask

StrBase.include

StrBase.index

StrBase.is_aligned

StrBase.is_mutable

StrBase.maxbit

StrBase.mmhash128

StrBase.msk_ascii_16

StrBase.msk_latin_16

StrBase.msk_lead

StrBase.msk_num2b_16

StrBase.next

StrBase.normalize

StrBase.nothing_sentinel

StrBase.output_utf8_2byte!

StrBase.output_utf8_3byte!

StrBase.output_utf8_4byte!

StrBase.pushchar!

StrBase.reverseind

StrBase.safe_copy

StrBase.set_codeunit!

StrBase.skipascii

StrBase.skipbmp

StrBase.splitarr

StrBase.swappedtype

StrBase.to_ascii

StrBase.unsafe_check_string — Function.

Validates and calculates number of characters in a UTF-8,UTF-16 or UTF-32 encoded vector/string

Warning: this function does not check the bounds of the start or end positions Use check_string to make sure the bounds are checked

Input Arguments:

dat UTF-8 (Vector{UInt8}), UTF-16 (Vector{UInt16}) or UTF-32 (Vector{UInt32}, AbstractString) encoded string

pos start position

endpos end position

Keyword Arguments:

accept_long_null = false # Modified UTF-8 (\0 represented as b"\xc0\x80")

accept_surrogates = false # CESU-8

accept_long_char = false # Accept arbitrary long encodings

accept_invalids = false # Accept invalid sequences (to be replaced on conversion)

Returns:

(total characters, flags, 4-byte, 3-byte, 2-byte)

Throws:

StringError

StrBase.unsafe_str — Function.

Convert to a UniStr if valid Unicode, otherwise return a Text1Str

StrBase.utf16 — Function.

utf16(s)

Create a UTF-16 string from a byte array, array of UInt16, or any other string type. (Data must be valid UTF-16. Conversions of byte arrays check for a byte-order marker in the first two bytes, and do not include it in the resulting string.)

utf16(::Union{Ptr{UInt16}, Ptr{Int16}} [, length])

Create a string from the address of a NUL-terminated UTF-16 string. A copy is made; the pointer can be safely freed. If length is specified, the string does not have to be NUL-terminated.

StrBase.utf32 — Function.

utf32(s)

Create a UTF-32 string from a byte array, array of Char or UInt32, or any other string type. (Conversions of byte arrays check for a byte-order marker in the first four bytes, and do not include it in the resulting string.)

utf32(::Ptr{UInt32}, [, length])

Create a string from the address of a NUL-terminated UTF-32 string. A copy is made; the pointer can be safely freed. If length is specified, the string does not have to be NUL-terminated.

StrBase.utf8 — Function.

utf8(s)

Create a UTF-8 string from a byte array, array of UInt8, or any other string type. (Data must be valid UTF-8. Conversions of byte arrays check for a byte-order marker in the first two bytes, and do not include it in the resulting string.)

utf8(::Union{Ptr{UInt8}, Ptr{Int8}} [, length])

Create a string from the address of a NUL-terminated UTF-8 string. A copy is made; the pointer can be safely freed. If length is specified, the string does not have to be NUL-terminated.

StrBase.utf8proc_map

StrBase.utf_length