Docstrings

LocalDiskObjectStores.Bucket

LocalDiskObjectStores.LocalDiskObjectStore

LocalDiskObjectStores.LocalDiskObjectStores

LocalDiskObjectStores.Object

LocalDiskObjectStores.ObjectStore

LocalDiskObjectStores.ObjectStores

LocalDiskObjectStores.Permission

LocalDiskObjectStores._create! — Function.

Create bucket. If successful return nothing, else return an error message as a String.

Create object. If successful return nothing, else return an error message as a String.

LocalDiskObjectStores._delete! — Function.

Delete bucket. If successful return nothing, else return an error message as a String.

Delete object. If successful return nothing, else return an error message as a String.

LocalDiskObjectStores._isbucket

LocalDiskObjectStores._isobject

LocalDiskObjectStores._read — Function.

Read bucket. If successful return (true, value), else return (false, errormessage::String).

Read object. If successful return (true, value), else return (false, errormessage::String).

LocalDiskObjectStores.createbucket!

LocalDiskObjectStores.delete!

LocalDiskObjectStores.deletebucket!

LocalDiskObjectStores.eval

LocalDiskObjectStores.getindex

LocalDiskObjectStores.getpermission

LocalDiskObjectStores.haspermission

LocalDiskObjectStores.include

LocalDiskObjectStores.isbucket

LocalDiskObjectStores.isobject

LocalDiskObjectStores.listcontents

LocalDiskObjectStores.permissions_conflict

LocalDiskObjectStores.setexpiry!

LocalDiskObjectStores.setindex!

LocalDiskObjectStores.setpermission!