Docstrings

TreeViews.TreeViews

TreeViews.eval

TreeViews.hastreeview — Function.

hastreeview(x::T)::Bool

Called by a frontend to decide whether a tree view for type T is available. Defaults to false.

TreeViews.include

TreeViews.nodelabel — Function.

nodelabel(io::IO, x::T, i::Integer, mime::MIME"text/plain" = MIME"text/plain"())

Prints the label of x's i-th child to io. Like with Base.show there are also methods with mime::AbstractString and no mime argument at all (which falls back to MIME"text/plain"()). Please only overload the nodelabel(io::IO, x, i::Integer, mime::MIME) form.

TreeViews.numberofnodes — Function.

numberofnodes(x)

Number of direct descendents. Defaults to fieldcount(typeof(x)).

TreeViews.treelabel — Function.

treelabel(io::IO, x, mime = MIME"text/plain"())

Prints x's tree header to io. Like with Base.show there are also methods with mime::AbstractString and no mime argument at all (which falls back to MIME"text/plain"()). Please only overload the treelabel(io::IO, x, mime::MIME) form.

TreeViews.treenode — Function.

treenode(x::T, i::Integer)

Returns the i-th node of x, which is usually printed by the display frontend next to the corresponding treelabel.