Docstrings

ClassImbalance.ClassImbalance

ClassImbalance.RealOrMissing

ClassImbalance._smote

ClassImbalance.cases_needed

ClassImbalance.classlabel — Function.

classlabel(y)

Given a column from a DataFrames.DataFrame, this function returns the majority/minority class label.

ClassImbalance.column_ranges

ClassImbalance.dataframe_to_matrix

ClassImbalance.eval

ClassImbalance.factor_columns

ClassImbalance.factor_to_float

ClassImbalance.fill_diagonal!

ClassImbalance.float_to_factor

ClassImbalance.include

ClassImbalance.matrix_to_dataframe

ClassImbalance.ncol

ClassImbalance.numeric_columns

ClassImbalance.pct_needed

ClassImbalance.rose

ClassImbalance.rose_real

ClassImbalance.rose_sampling

ClassImbalance.rscale

ClassImbalance.smote — Function.

smote(X, y, k, under, over)

This function implements the SMOTE algorithm for generating synthetic cases to re-balance the proportion of positive and negative observations. The pct_under and pct_over parameters control the proportion of under-sampling of the majority class and the proportion of over-sampling the minority class. Note that pct_under controls undersampling by selecting pctunder/100 observations for each _newly created minority class observation. The value of k allows us determine who is considered a "neighbor" when generating synthetic cases.

ClassImbalance.smote_obs