Docstrings

TSVD.AtA

TSVD.TSVD

TSVD._teig

TSVD._tsvd

TSVD.biLanczosIterations

TSVD.eval

TSVD.include

TSVD.lanczosIterations

TSVD.qr!

TSVD.teig

TSVD.tsvd — Function.

tsvd(A, nvals = 1; [maxiter, initvec, tolconv, tolreorth, debug])

Computes the truncated singular value decomposition (TSVD) by Lanczos bidiagonalization of the operator A. The Lanczos vectors are partially orthogonalized as described in

R. M. Larsen, Lanczos bidiagonalization with partial reorthogonalization, Department of Computer Science, Aarhus University, Technical report, DAIMI PB-357, September 1998.

Positional arguments:

A: Anything that supports the in place update operations

mul!(y::AbstractVector, A, x::AbstractVector, α::Number, β::Number)

and

mul!(y::AbstractVector, A::Adjoint, x::AbstractVector, α::Number, β::Number)

corresponding to the operations y := α*op(A)*x + β*y where op can be either the identity or the conjugate transpose of A. If the initvec argument is not supplied then it is furthermore required that A supports eltype and size.

nvals: The number of singular values and vectors to compute. Default is one (the largest).

Keyword arguments:

maxiter: The maximum number of iterations of the Lanczos bidiagonalization. Default is 1000, but usually much fewer iterations are needed.

initvec: Initial U vector for the Lanczos procesdure. Default is a vector of Gaussian random variates. The length and eltype of the initvec will control the size and element types of the basis vectors in U and V.

tolconv: Relative convergence criterion for the singular values. Default is sqrt(eps(real(eltype(A)))).

tolreorth: Absolute tolerance for the inner product of the Lanczos vectors as measured by the ω recurrence. Default is sqrt(eps(real(eltype(initvec)))). 0.0 and Inf correspond to complete and no reorthogonalization respectively.

debug: Boolean flag for printing debug information

Output:

The output of the procesure it the truple tuple (U, s, V)

U: size(A, 1) times nvals matrix of left singular vectors.

s: Vector of length nvals of the singular values of A.

V: size(A, 2) times nvals matrix of right singular vectors.

Examples

julia> A = matrixdepot("Rucci/Rucci1", :r)
1977885×109900 SparseMatrixCSC{Float64,Int64} with 7791168 stored entries:
  [1      ,       1]  =  0.167285
  [2      ,       1]  =  0.111776
  [3      ,       1]  =  0.0746865
  [208    ,       1]  =  0.0703943
  [209    ,       1]  =  0.0300765
  [210    ,       1]  =  0.0128504
  [415    ,       1]  =  0.00777096
  ⋮
  [1977263,  109900]  =  0.0618182
  [1977264,  109900]  =  0.151233
  [1977675,  109900]  =  0.0869268
  [1977677,  109900]  =  0.139329
  [1977678,  109900]  =  0.223321
  [1977882,  109900]  =  0.224977
  [1977884,  109900]  =  0.268216
  [1977885,  109900]  =  0.319764

julia> U, s, V = tsvd(A, 5);

julia> s
5-element Array{Float64,1}:
 7.06874
 6.98525
 6.96246
 6.88948
 6.86255

TSVD.tsvd2