Docstrings

GtkUtilities.AbstractState

GtkUtilities.GtkUtilities — Module.

Summary of features in GtkUtilities:

rubberband_start: initiate rubber band selection

guidata: associate user data with on-screen elements

Each of these has detailed help available, e.g., ?guidata.

GtkUtilities.GuiData

GtkUtilities.PanZoom.Interval — Type.

An Interval is a (min,max) pair. It is the one-dimensional analog of a BoundingBox.

GtkUtilities.Link

GtkUtilities.PanZoom

GtkUtilities.RubberBands

GtkUtilities.State

GtkUtilities.disconnect

GtkUtilities.eval

GtkUtilities.PanZoom.fullview — Function.

iv = fullview(limits) returns an Interval iv that encompases the full view as permitted by limits.

If limits == nothing, then fullview returns nothing.

The simplest effectual limits object is another Interval representing the "whole canvas" along the chosen axis. If you need more sophisticated behavior, you can extend this function to work with custom types of limits objects.

GtkUtilities.get

GtkUtilities.GuiData.guidata — Constant.

Given a widget (Button, Canvas, Window, etc.) or other graphical object w, a value val can be associated with ("stored in") w using

guidata[w, :name] = val

where :name is the name (a Symbol) you've assigned to val for the purposes of storage.

The value can be retrieved with

val = guidata[w, :name]

Here are some other things you can do with guidata:

alldata = guidata[w]           # fetch all data associated with w
val = get(guidata, (w,:name), default)   # returns default if :name not defined
delete!(guidata, (w,:name))    # deletes the value associated with :name
delete!(guidata, w)            # deletes all data associated with w

Example:

    c = Canvas()
    panzoom(c, (0,1), (0,1))
    xview = guidata[c, :xview]

Note that if :name corresponds to a State object, guidata will get/set the value of the State object, transparently stripping the wrapper itself. This means that you change the value without destroying any links you have set up. If you need the state object itself, use the raw keyword:

    state = getindex(guidata, c, :name; raw=true)     # retrieves a state object
    setindex!(guidata, newstate, c, :name; raw=true)  # replaces the old state object

GtkUtilities.id

GtkUtilities.image_surface

GtkUtilities.image_to_cairo

GtkUtilities.include

GtkUtilities.PanZoom.interior — Function.

ivnew = interior(iv, limits) returns a new version of iv, an Interval, which is inside the region allowd by limits. One should prefer "shifting" iv over "shrinking" iv (if possible, the width of iv should be preserved).

If limits == nothing, then iv is unconstrained and ivnew == iv.

The simplest effectual limits object is another Interval representing the full view interval across the chosen axis. If you need more sophisticated behavior, you can extend this function to work with custom types of limits objects.

GtkUtilities.lEntry

GtkUtilities.lLabel

GtkUtilities.lScale

GtkUtilities.Link.link — Function.

w_linked = link(state, widget) links the value of the user-interface element widget to the value of the AbstractState state. The two will henceforth be synchronized: calling get(state) or get(w_linked) returns the current value, and set!(state, val) or set!(w_linked, val) will change the value for all mutually-linked objects.

link(state, c), where c is a Canvas, makes c a listener for state. There is no return value.

GtkUtilities.PanZoom.panzoom — Function.

panzoom(c)
panzoom(c, xviewlimits, yviewlimits)
panzoom(c, xviewlimits, yviewlimits, xview, yview)
panzoom(c2, c1)

sets up the Canvas c for panning and zooming. The arguments may be 2-tuples, 2-vectors, Intervals, or nothing.

panzoom creates the :view[x|y], :view[x|y]limits properties of c:

:xview, :yview are two AbstractStates (for horizontal and vertical, respectively), each holding an Interval specifying the current "view" limits. This might be the entire area, or it might be a subregion due to a previous zoom event.

:xviewlimits, :yviewlimits encode the maximum allowable viewing region; in most cases these will also be State{Interval}s, but any object that supports interior and fullview may be used. Use nothing to indicate unlimited range.

If c is the only argument to panzoom, then the current user-coordinate limits of c are used. Note that this invocation works only if the Canvas has been drawn at least once; if that is not the case, you need to specify the limits manually.

panzoom(c2, c1) sets canvas c2 to share pan/zoom state with canvas c1. Panning or zooming in either one will cause the same action in the other.

GtkUtilities.PanZoom.panzoom_key — Function.

id = panzoom_key(c; kwargs...) initializes panning- and zooming-by-keypress for a canvas c. c is expected to have the four guidata properties described in panzoom.

You can configure the keys through keyword arguments. The default settings are shown below. The first entry is the key, the second a modifier (like the SHIFT key); 0 means no modifier.

    panleft      = (GDK_KEY_Left,0),
    panright     = (GDK_KEY_Right,0),
    panup        = (GDK_KEY_Up,0),
    pandown      = (GDK_KEY_Down,0),
    panleft_big  = (GDK_KEY_Left,SHIFT),
    panright_big = (GDK_KEY_Right,SHIFT),
    panup_big    = (GDK_KEY_Up,SHIFT),
    pandown_big  = (GDK_KEY_Down,SHIFT),
    xpanflip     = false,
    ypanflip     = false
    zoomin       = (GDK_KEY_Up,  CONTROL)
    zoomout      = (GDK_KEY_Down,CONTROL)

"Regular" panning motions correspond to 10% of the view region; "big" panning motions are 100% of the view region, and thus jump by one whole view area. The constants are defined in Gtk.GConstants and the modifiers in Gtk.GConstants.GdkModifierType.

The returned id can be disabled or enabled via signal_handler_block and signal_handler_unblock, respectively, or removed with signal_handler_disconnect.

Example:

    c = Canvas()
    panzoom(c, (0,1), (0,1))
    id = panzoom_key(c)

The draw method for c should take account of :viewbb.

GtkUtilities.PanZoom.panzoom_mouse — Function.

panzoom_mouse(c; kwargs...) initializes panning-by-mouse-scroll and mouse control over zooming for a canvas c.

zooming or panning (along either x or y) is selected by modifier keys, which are configurable through keyword arguments. The default settings are:

    # Panning
    xpan      = SHIFT     # hold down the shift key
    ypan      = 0
    xpanflip  = false
    ypanflip  = false
    # Zooming
    zoom      = CONTROL     # hold down the ctrl-key while scrolling
    focus     = :pointer    # zoom around the position under the mouse pointer
    factor    = 2.0
    initiate  = BUTTON_PRESS # start a rubberband selection for zoom
    reset     = DOUBLE_BUTTON_PRESS    # go back to original limits

where 0 means no modifier. SHIFT, CONTROL, BUTTONPRESS, and DOUBLEBUTTON_PRESS are defined in Gtk.GConstants.GdkModifierType.

The focus keyword controls how the zooming progresses as you scroll the mouse wheel. :pointer means that whatever feature of the canvas is under the pointer will stay there as you zoom in or out. The other choice, :center, keeps the canvas centered on its current location. These behaviors are subject to modification by the canvas' :viewlimits data.

An additional keyword is user_to_data, for which you may supply a function

    user_to_data_fcn(c, x, y) -> (datax, datay)

that converts canvas user-coordinates to "data coordinates" before setting the values of :xview and :yview.

For important additional information, see panzoom_key. To disable mouse panning and zooming, use

    pop!((c.mouse, :scroll))
    pop!((c.mouse, :button1press))

Example:

    c = Canvas()
    panzoom(c, (0,1), (0,1))
    panzoom_mouse(c)

GtkUtilities.RubberBands.rubberband_start — Function.

rubberband_start(c, x, y, callback_done; minpixels=2) starts a rubber-band selection on Canvas c at position (x,y). When the user releases the mouse button, the callback function callback_done(c, bb) is run, where bb is the BoundingBox of the selected region. To reduce the likelihood that clicks used to raise windows will result in rubber banding, the callback is not executed unless the user drags the mouse by at least minpixels pixels (default value 2).

Example:

c.mouse.button1press = (widget, event) -> begin
    if event.event_type == Gtk.GdkEventType.BUTTON_PRESS
        GtkUtilities.rubberband_start(c, event.x, event.y, (c, bb) -> @show bb)
    end
end

would set up a Canvas so that rubberband selection starts when the user clicks the mouse; when the button is released, it displays the bounding box of the selection region.

GtkUtilities.Link.set! — Function.

set(w, val) sets the value of the linked widget w and fires the callback, thereby updating all other linked widgets.

GtkUtilities.set_coordinates

GtkUtilities.Link.set_quietly! — Function.

set_quietly!(w, val) sets the value of the linked widget w without firing the callback.

GtkUtilities.GuiData.trigger — Function.

trigger(widgets, symbols) is used when you want to set several state variables simultaneously, but don't want to refresh the screen more frequently than necessary. You can set the .value parameter of the state variables directly, then call trigger to synchronize the GUI.

GtkUtilities.widget