Docstrings

ROCAnalysis.AUC — Function.

AUC(::roc; pfa=1.0, pmiss=1.0, normalize=true) computes the traditional Area Under the Curve with a sense AUC → 1 indicating better performance.

Optional parameters pfa or pmiss limit integration over only part of the ROC curve. normalize indicates comparing the partial ROC to the trivial ROC.

You can also call this function as AUC(targets::Vector, nontargets::Vector; kwargs...)

ROCAnalysis.ArrayOrReal

ROCAnalysis.DCF — Type.

DCF is a type that holds one or more cost functions with which the performance of a two-class classifier can be assessed. Any of the fields can be either a scalar or an array.

DataFrames.DataFrame — Type.

Dataframe(::Roc) converts a Roc object in a datframe with columns:

pfa the false alarm rate

pmiss the miss rate

thres the thereshold, separating this line's pfa and pmiss from the next

chull indicating if this point is on the convex hull of the ROC curve

llr the optimal log-likelihood-ratio score for all data points contributing to the ROC line segment from this line to the next

ROCAnalysis.ROCAnalysis

ROCAnalysis.Roc — Type.

Roc is a type that stores the essential performance information that can be extracted from a set of supervised trials, i.e., target and non-target scores from a two-class classifier. Apart from the (minimalized) arrays for probabilities of false-alarm and miss–-the coordinates of the ROC curve–-, they are the threshold for these, a boolean whether-or-not this point lies on the convex hull, and the associated optimal log-likelihood-ratio associated to the line segment.

ROCAnalysis.TNT — Type.

TNT(tar, non) is a type that holds target and non-target scores in a test for a two-class classifier.

ROCAnalysis.apeplot

ROCAnalysis.applyfactor

ROCAnalysis.auc — Function.

AUC(::roc; pfa=1.0, pmiss=1.0, normalize=true) computes the Area Under the Curve with a sense auc → 0 indicating better performance.

Optional parameters pfa or pmiss limit integration over only part of the ROC curve. normalize indicates comparing the partial ROC to the trivial ROC.

ROCAnalysis.ber — Function.

ber(), the Bayes Error Rate. This computes the expected error rate given a set of supervised trials, the log-likelihood-ratio scores tar and non, and a threshold based on a given a cost function specified by its prior log-odds (see plo()).

If the scores tar and non are well-calibrated log-likelihood-ratios, the optimum decision threshold that minimizes the decision cost function dcf is surprisingly simple,

θ = -plo(dcf).

This function computes the actual cost for such decisions, given the test scores.

Arguments:

tar::Vector, non::Vector, plo: target and non-target llr

scores, and prior log odds. Here, plo may be a vector, resulting in multiple Bayes error rates.

r::Roc, plo: a Roc structure computed with roc(;collapse=false).

It is mandatory that the Roc object is not collapsed, because the actual cost may occur for a threshold between two "corner" points of the ROC curve. A Roc structure is useful for computing the minimum Bayes Error Rate for multiple cost functions–-in this case the Roc may be collapsed.

; thres: an optional threshold, overriding the theoretical

optimum described above. This can be used if the scores are not correctly calibrated.

ROCAnalysis.ber_famiss

ROCAnalysis.binscores

ROCAnalysis.binsearch

ROCAnalysis.changepoints

ROCAnalysis.chllr

ROCAnalysis.cllr — Function.

cllr(tar, non) computes the cost of the log-likelihood-ratio, for log-likelihood-ratio (llr) scores tar and non (target scores and non-target scores). Target an non-target llr's are those for which the numerator H1 and denominator H2 hypothesis are actually true in the log-likelihood-ratio

llr(x) = p(x|H1) / p(x|H2)

where x is the test data for a system trying to disciminate between H1 and H2 given x.

The Cllr of a perfect system, assigning a llr of +Inf t otarget scores and -Inf to non-target scores is 0, for a system that is indifferent, producing a llr=0 for every input x the Cllr=1. Please note that, for badly calibrated systems, Cllr>1. The units of Cllr are measured in bits, and cllr can be seen as the average amount of information per trial that is not extracted from the data.

Cllr measures the calibration as well as discrimination ability of a system. Discrimination entails the ability to produces target scores that are, typically, much higher than non-target scores. Calibration entails that for every individual x an optimal Bayes decision can be made if a cost function is known.

ROCAnalysis.costfactor

ROCAnalysis.crossing

ROCAnalysis.dcf — Function.

dcf(). Compute the classical decision cost function:

dcf = ptar Cmiss pmiss + (1-ptar) Cfa pfa.

The relative value of the cost function is only determined by the ratio

oeff = (ptar/(1-ptar) (Cmiss/Cfa).

Therefore, this function is computed using ber() and scaling accordingly. Arguments:

tar, non: target and non-target scores

tnt::TNT: a TNT object containing target and non-target scores.

r::Roc: a Roc object, the result of roc()

; d::DCF: a decision cost function, default getdcf(). This can be a vector or DCFs.

; thres: the threshold used to make decisions, default -plo(d)

; norm: are the costs normalized to a trivial system deciding using the prior only, default: false

ROCAnalysis.detplot

ROCAnalysis.eer — Function.

eer(tar, non) computes a simple approximation to the Equal Error Rate, the intersection of the ROC with the line y=x, i.e., the error rate at which the false positive rate and the false negative rate are approximately equal.

For a more consistent interpretation and implementation of the EER, please consider eerch(), the Equal Error Rate - Convex Hull computation.

ROCAnalysis.eer_experimental

ROCAnalysis.eer_so

ROCAnalysis.eer_sorted

ROCAnalysis.eerch — Function.

eerch() computes the Equal Error Rate using the Convex Hull interpretation. This computes the error rate at which the convex hull of the ROC curve crosses the y=x diagonal in the ROC.

This value is less sensitive to trial sampling at low number of trials, and has the interpretation of `the maximum while vary prior of the minimum decision costs'–-a useful quantity in decision cost function analysis of a two class classification system.

Arguments:

r::Roc: an object of type Roc, the result of roc()

tar, non: Vectors of target and non-target scores

pfa, pmiss, ch: Vectors of the false positive and false negative rate, and an array indicating

the membership of the (pfa, pmiss) point on the convex hull. These points are found by roc().

ROCAnalysis.eval

ROCAnalysis.getdcf — Function.

getdcf() retrieves the current valua of the DCF parameters.

ROCAnalysis.global_dcf

ROCAnalysis.include

ROCAnalysis.isleft

ROCAnalysis.llrplot

ROCAnalysis.logit

ROCAnalysis.minber — Function.

minber(): The minimum Byes Error Rate for a set of socres, found after an optimal score-to-llr (log-likelihood-ratio) transformation. The optimal transformation corresponds to the convex hull of the ROC, where the optimal llr values correspont to the negative log of the slope of the appropriate line segment spanning the ROC. This is equivalent to running the pool-adjacent-violators algorithm on the test data.

In order to compute minber() for multiple cost functions, as in a Normalized Bayes Error Rate plot, it is advantageous to compute a Roc object once.

Arguments: see ber().

ROCAnalysis.minber_famiss

ROCAnalysis.mincllr — Function.

mincllr(tar, non) computes the cllr of target and non-target scores after an optimal transformation of the data. This allows for measuring the disrimination performance of a system in the units of cllr, which are bits.

ROCAnalysis.mindcf — Function.

mindcf() computes the minimum costs of the classical decision cost function (see dcf()) that can be obtained by varying the threshold. Arguments are the same as for dcf(), except that there is no threshold.

This function uses minber() and scales accordingly.

ROCAnalysis.missing

ROCAnalysis.nbeplot

ROCAnalysis.normfactor

ROCAnalysis.oeff — Function.

oeff(), the effective prior odds given cost function parameters. Arguments are either:

ptar, cfa, cmiss: Scalars or Vectors of the prior of a target, the cost of a false positive and the cost of a false negative

dcf::DCF: a DCF object containing ptar, cfa and cmiss.

ROCAnalysis.optllr

ROCAnalysis.pav

ROCAnalysis.peff — Function.

peff(), the effective prior given cost function parameters. Arguments are either:

ptar, cfa, cmiss: Scalars or Vectors of the prior of a target, the cost of a false positive and the cost of a false negative

dcf::DCF: a DCF object containing ptar, cfa and cmiss.

ROCAnalysis.plo — Function.

plo(), the prior log-odds given cost function parameters. Arguments are either:

ptar, cfa, cmiss: Scalars or Vectors of the prior of a target, the cost of a false positive and the cost of a false negative

dcf::DCF: a DCF object containing ptar, cfa and cmiss.

ROCAnalysis.plot

ROCAnalysis.pnorm

ROCAnalysis.qnorm

ROCAnalysis.roc — Function.

roc(tar, non; laplace, collapse) computes the essential statistics for evaluation of the performance of a two-class classifier. The true class of the scores is encoded in the array in which they appear, i.e., target or non-target scores.

The results are given in a type Roc, containing the receiver-operating-characteristics (ROC) data for this test.

Optional arguments:

collapse=true, indicating whether the resulting araays for the probability of

false positives (alarms) and negatives (misses) should be collapsed, i.e., consecutive points on the ROC that have the same false positive or false negatove rate are removed, retaining only the `corner' points of the ROC.

laplace=false, indicating whether two additional data points at either end of the score scale

should be added to the target an dnon-target scores. This corresponds to the Laplace prior, and has a result of limiting the magnitude of the optimal log-likelihood-ratio associated with the line segments of the convex hull of the ROC.

ROCAnalysis.rochull

ROCAnalysis.rocplot

ROCAnalysis.setdcf — Function.

setdcf() allows a global decision cost function to be set. This DCF can be used in dcf() and mindcf(). The decision cost function is defined as

dcf = ptar Cmiss pmiss + (1-ptar) Cfa pfa.

and is defined by the parameters ptar (the target prior), Cfa (the cost of a false positive) and Cmiss (the cost of a false negative. Arguments:

ptar. The target prior, default 0.5.

cfa. The cost of a false positive (false alarm), default 1

cmiss. The cost of a false negative (miss), default 1.

Multiple simultaneous cost functions can be set by specifying any, or a combination, of these parameters as Vectors.

The current values of the parameters of the DCF can be found using getdcf().

ROCAnalysis.sigmoid

ROCAnalysis.softplus

ROCAnalysis.sortscores