Docstrings

Hive.@logmsg

Hive.BIT_MASKS

Hive.Col

Hive.ColValue

Hive.ColumnChunksIterator

Hive.DEFAULT_FETCH_SIZE — Constant.

number of records to fetch with every fetchnext

Hive.HS2

Hive.Hive

Hive.HiveAuth — Type.

Authentication mechanisms Only SASL-Plain supported for now

Hive.HiveAuthSASLPlain

Hive.HiveConn — Type.

HiveConn holds the thrift connection and protocol objects. It also holds the hive session handle for this connection.

Hive.HiveSession — Type.

HiveSession holds a connection and session status

Hive.InfoType — Constant.

Info Type

Hive.JCONV

Hive.JTYPES

Hive.Pattern — Type.

A search pattern used in the below methods can have:

'_': Any single character.

'%': Any sequence of zero or more characters.

'': Escape character used to include special characters,

e.g. '_', '%', ''. If a '' precedes a non-special character it has no special meaning and is interpreted literally.

Hive.PendingResult

Hive.READY_STATUS

Hive.RESULT_STATUS

Hive.RecordIterator

Hive.Result

Hive.ResultSet

Hive.RowCount

Hive.SCHEMA_CACHE

Hive.Tabular

Hive.TabularIterator

Hive.bitset_to_bools

Hive.cancel

Hive.catalogs — Function.

Returns the list of catalogs (databases). Results are ordered by TABLE_CATALOG.

Hive.check_status

Hive.close

Hive.colconvfns

Hive.coltypes

Hive.columnchunk

Hive.columnchunks

Hive.columns — Function.

Returns a list of columns in the specified tables. Optional parameter catalog must contain a full catalog name. Optional parameters schema_pattern, table_pattern and column_pattern can contain a search pattern. Result Set Columns are the same as those for the ODBC CLIColumns function.

Hive.compactshow

Hive.data_with_missings

Hive.eof

Hive.eval

Hive.execute — Function.

Execute a statement. Statement to be executed can be DML, DDL, SET, etc. Optional config parameter can have additional keyword parameters that will be passed as configuration properties that are overlayed on top of the the existing session configuration before this statement is executed. They apply to this statement only and are not permanent. When async is true, execution is asynchronous and a PendingResult may be returned. If the returned value is a PendingResult: - Caller must call isready on a PendingResult instance to check for completion. - Once ready, calling result on it returns ResultSet (the same PendingResult instance is returned if it is still not ready)

Hive.fetchfirst

Hive.fetchnext

Hive.fetchsize

Hive.fetchsize!

Hive.functions — Function.

Returns a list of functions supported by the data source. Catalog name must match the catalog name as it is stored in the database; "" retrieves those without a catalog; null means that the catalog name should not be used to narrow the search. Schema name pattern must match the schema name as it is stored in the database; "" retrieves those without a schema; null means that the schema name should not be used to narrow the search. Function name pattern must match the function name as it is stored in the database. The behavior of this function matches java.sql.DatabaseMetaData.getFunctions().

Hive.get_info

Hive.hasresult

Hive.include

Hive.isready

Hive.iterate

Hive.iteratorsize

Hive.julia_conv

Hive.julia_type

Hive.julia_type_name

Hive.records

Hive.response_to_resultset

Hive.response_to_tabular

Hive.result

Hive.schema

Hive.schemas — Function.

Retrieves the schema names available in this database. The results are ordered by TABLECATALOG and TABLESCHEM.

Hive.show

Hive.show_table

Hive.show_table_meta

Hive.sqlerror

Hive.status

Hive.tables — Function.

Returns a list of tables with catalog, schema, and table type information. Results are ordered by TABLETYPE, TABLECAT, TABLESCHEM, and TABLENAME

Hive.tabletypes — Function.

Returns the table types available in this database. The results are ordered by table type.

Hive.tabular

Hive.tabulars

Hive.tobigfloat

Hive.tochar

Hive.todate

Hive.todatetime

Hive.tofloat32

Hive.tofloat64

Hive.toint32

Hive.toint64

Hive.toint8