Docstrings

SimpleGraphs.AbstractSimpleGraph — Type.

AbstractSimpleGraph is a parent class for SimpleGraph and SimpleDigraph.

SimpleGraphs.BuckyBall — Function.

BuckyBall() returns the Bucky ball graph.

SimpleGraphs.Complete — Function.

Complete(n) returns a complete graph with n vertices 1:n.

Complete(n,m) returns a complete bipartite graph with n vertices in one part and m vertices in the other.

Complete([n1,n2,...,nt]) returns a complete multipartite graph with parts of size n1, n2, ..., nt.

SimpleGraphs.Cube — Function.

Cube(n) creates the n-dimensional cube graph. This graph has 2^n vertices named by all possible length-n strings of 0s and 1s. Two vertices are adjacent iff they differ in exactly one position.

SimpleGraphs.Cycle — Function.

Cycle(n) creates a cycle with vertex set 1:n.

SimpleGraphs.DEFAULT_MARKER_SIZE

SimpleGraphs.DFS — Function.

perform a depth first search on graph G starting at vertex v

SimpleGraphs.DirectedComplete — Function.

DirectedComplete(n) creates a directed complete graph with all possible edges (including a loop at each vertex). Use DirectedComplete(n,false) to supress the creation of loops.

SimpleGraphs.DirectedCycle — Function.

DirectedCycle(n) creates a directed cycles with vertices 1:n.

SimpleGraphs.DirectedPath — Function.

DirectedPath(n) creates a directed cycles with vertices 1:n.

SimpleGraphs.Dodecahedron — Function.

Dodecahedron() creates the dodecahedron SimpleGraph.

SimpleGraphs.Frucht — Function.

Frucht() returns the Frucht graph: A 12-vertex, 3-regular graph with no non-nontrivial automorphisms.

SimpleGraphs.GraphEmbedding — Type.

GraphEmbedding(G) creates a drawing of the SimpleGraph G. It is given a circular embedding.

Note: Direct usage of GraphEmbedding is deprecated!

GraphEmbedding(G,d) creates a drawing using the Dict d to specify the vertex locations. d should map vertices of G to Vectors that specify x,y-coordinates.

SimpleGraphs.Grid — Function.

Grid(n,m) creates an n-by-m grid graph. For other grids, we suggest Path(n1)*Path(n2)*Path(n3) optionally wrapped in relabel. See also: Cube.

SimpleGraphs.Hoffman — Function.

Hoffman() creates the Hoffman graph which is cospectral with, but not isomorphic to, Cube(4).

SimpleGraphs.HoffmanSingleton — Function.

HoffmanSingleton() creates the Hoffman-Singleton graph. This is a 7-regular graph whose diameter is 2 and whose girth is 5.

SimpleGraphs.Icosahedron — Function.

Icosahedron() creates the icosahedron SimpleGraph.

SimpleGraphs.IntDigraph — Function.

IntDigraph() creates a new directed graph with vertices of type Int64.

IntDigraph(n) prepopulates the vertex set with vertices 1:n.

SimpleGraphs.IntGraph — Function.

IntGraph() creates a new SimpleGraph whose vertices are of type Int. Called as IntGraph(n::Int) prepopulates the vertex set with vertices 1:n.

IntGraph(A) where A is an adjacency matrix creates a graph for which A is the adjacency matrix.

SimpleGraphs.Johnson — Function.

Johnson(n,k) creates the Johnson graph whose vertices and the k-element subsets of {1,2,...,n}. Vertices v and w are adjacent if their intersection has size k-1.

SimpleGraphs.Kneser — Function.

Kneser(n,m) creates the Kneser graph whose vertices are all the m-element subsets of 1:n in which two vertices are adjacent iff they are disjoint.

SimpleGraphs.Knight — Function.

Knight(r::Int=8,c::Int=8) creates a Knight's Moves graph on a r-by-c grid. That is, the vertices of this graph are the squares of an r-by-c chess board. Two vertices are adjacent if a Knight can go from one of these squares to the other in a single move.

SimpleGraphs.LZ — Function.

Computes L^Z defined in (5) of the Reference

Input: Z: current layout (coordinates) d: Ideal distances (default: all 1) w: weights (default: d.^-2)

SimpleGraphs.NE — Function.

NE(G) returns the number of edges in G.

SimpleGraphs.NV — Function.

NV(G) returns the number of vertices in G.

SimpleGraphs.Octahedron — Function.

Octahedron() creates the octaahedron SimpleGraph.

SimpleGraphs.Paley — Function.

Paley(p) creates the Paley graph with p vertices named 0:(p-1). Here p must be a prime with p%4==1. Vertices u and v are adjacent iff u-v is a quadratic residue (perfect square) modulo p.

SimpleGraphs.Path — Function.

Path(n) creates a path graph with n vertices named 1:n.

Path(array) creates a path graph with vertices array[1], array[2], etc.

SimpleGraphs.Petersen — Function.

Petersen() returns the Petersen graph. The vertices are labeled as the 2-element subsets of 1:5. Wrap in relabel to have vertices named 1:10. See also: Kneser.

SimpleGraphs.RandomDigraph — Function.

RandomDigraph(n,p) creates an Erdos-Renyi style random directed graph with vertices 1:n and edge probability p (equal to 0.5 by default). The possible edges (u,v) and (v,u) are independent. No loops are created. To also create loops (each with probability p) use RandomDigraph(n,p,true).

SimpleGraphs.RandomGraph — Function.

RandomGraph(n,p=0.5) creates an Erdos-Renyi random graph with n vertices and edge probability p.

SimpleGraphs.RandomRegular — Function.

RandomRegular(n,d) creates a random d-regular graph on n vertices. This can take a while especially if the arguments are large. Call with an optional third argument to activate verbose progress reports: RandomRegular(n,p,true).

SimpleGraphs.RandomRegularBuilder

SimpleGraphs.RandomSBM — Function.

RandomSBM(bmap,pmat) creates a random stochastic block model random graph. The vector bmap is a list of n positive integers giving the block number of vertices 1:n. The i,j-entry of the matrix pmat gives the probability of an edge from a vertex in block i to a vertex in block j.

RandomSBM(n,pvec,pmat) creates such a graph with n vertices. The vector pvec gives the probabilities that vertices fall into a given block.

SimpleGraphs.RandomTournament — Function.

RandomTournament(n) creates a random tournament with vertex set 1:n. This is equivalent to randomly assigning a direction to every edge of a simple complete graph.

SimpleGraphs.RandomTree — Function.

RandomTree(n) creates a random tree on n vertices each with probability 1/n^(n-2).

SimpleGraphs.ShiftDigraph — Function.

ShiftDigraph(alphabet,n) creates a SimpleDigraph whose vertices are all length-n tuples of the elements in alphabet (which can be an array such as [0,1] or a string such as "abc"). An edge from u to v corresponds to an element dropped from the first position in u and another element added to the end yielding v. For example, in ShiftDigraph([0,1],5) there are two edges leaving vertex (0,1,0,1,1); one goes to (1,0,1,1,0) and the other to (1,0,1,1,1).

SimpleGraphs.SimpleDigraph — Type.

SimpleDigraph() creates a new directed graph with vertices of Any type. This can be restricted to vertics of type T with SimpleDigraph{T}().

SimpleGraphs.SimpleGraph — Type.

The SimpleGraph type represents a simple graph; that is, an undirected graph with no loops and no multiple edges.

Use SimpleGraph() to create a new graph in which the vertices may be Any type. Use SimpleGraph{T}() to create a new graph in which the vertices are of type T. See IntGraph and StringGraph as special cases.

SimpleGraphs.SimpleGraphs

SimpleGraphs.Spindle — Function.

Spindle() returns the Moser spindle graph. This is a seven-vertex unit distance graph with chromatic number equal to 4.

SimpleGraphs.StringDigraph — Function.

StringDigraph() creates a new directe graph with vertices of type String.

SimpleGraphs.StringGraph — Function.

A StringGraph is a SimpleGraph whose vertices are of type String.

When constructed with StringGraph() creates an empty SimpleGraph{String}.

When invoked as StringGraph(file::AbstractString) opens the named file for parsing as a graph. Each line of the file should contain one or two tokens separated by white space. If the line contains a single token, we add that token as a vertex. If the line contains two (or more) tokens, then the first two tokens are taken as vertex names and (assuming they are different) the corresponding edge is created. Any extra tokens on the line are ignored. Lines that begin with a # are ignored.

SimpleGraphs.Tetrahedron — Function.

Tetrahedron() creates the tetrahedron SimpleGraph.

SimpleGraphs.TorusDigraph — Function.

function for creating a Torus Graph

SimpleGraphs.Wheel — Function.

Wheel(n) creates a wheel graph with n vertices. That is, a cycle with n-1 vertices 1:(n-1) all adjacent to a common single vertex, n.

SimpleGraphs._checksquare

SimpleGraphs.add! — Function.

add!(G,v) adds a new vertex v to the graph.

add!(G,v,w) adds a new edge (v,w) to the graph. If one (or both) of those vertices is not already in the graph, it is added to the vertex set.

SimpleGraphs.add_dict!

SimpleGraphs.add_edge_matrix!

SimpleGraphs.add_edges! — Function.

add_edges!(G,edge_table) adds edges to the graph G. The argument edge_table is an m-by-2 array of vertices. Each row of the table represents an edge to add to G. Returns the number of edges successfully added to G.

This works both when G is a SimpleGraph and when G is a SimpleDigraph.

SimpleGraphs.add_key!

SimpleGraphs.add_key_value!

SimpleGraphs.add_value!

SimpleGraphs.adjacency — Function.

adjacency(G) returns the adjacency matrix of G.

Note: If the vertices can be sorted by sort, then the first row of the adjacency matrix corresponds to the first vertex (in order) in G and so forth. However, if the vertices are not sortable in this way, the mapping between vertices and rows/columns of the matrix is unpredictable.

Base.adjoint — Function.

G' is equivalent to complement(G).

SimpleGraphs.all_tuples — Function.

all_tuples(alphabet, n) creates an iterator that produces all length-n tuples of distinct elements in alphabet.

SimpleGraphs.allow_loops! — Function.

allow_loops!(G) enables G to have loops`.

SimpleGraphs.bipartition — Function.

bipartition(G) creates a bipartition of the graph (or returns an error if the graph is not bipartite. Output is a Partition.

SimpleGraphs.bisect — Function.

bisect(G::SimpleGraph) partitions the vertex set of G using the eigenvector associated with the second smallest eigenvalue of the graph's Laplacian matrix (called x below).

This can be invoked as follows:

bisect(G,"user",pivot) splits the vertices v depending on x[v] >= pivot vs. x[v] < pivot.

bisect(G,"zero") is the same as bisect(G,"user", 0.0).

bisect(G,"median") is equivalent to bisect(G,"user",m) where m is the median value of x.

bisect(G,"equal") creates a partition in which the two parts have sizes the differ by at most 1.

A plain call to bisect(G) is equivalent to bisect(G,"zero") (which is the same as bisect(G,"user", 0.0)).

SimpleGraphs.bounds

SimpleGraphs.cache_check — Function.

cache_check(G,item) checks if the symbol item is a valid key.

SimpleGraphs.cache_clear — Function.

cache_clear(G) clears all items in G's cache.

cache_clear(G,item) clears just that item.

SimpleGraphs.cache_off — Function.

cache_off(G) deactivates cache checking. You may also wish to call cache_clear to recover storage.

See also: cache_on.

SimpleGraphs.cache_on — Function.

cache_on(G) activates results caching for this graph. See also: cache_off.

SimpleGraphs.cache_recall — Function.

cache_recall(G,item) retreives the value associated with item.

WARNING: No check is done to see if this value is defined. Be sure to use cache_check first!

SimpleGraphs.cache_save — Function.

cache_save(G,item,value) saves value associated with the symbol (key) item in the cache for this graph.

SimpleGraphs.cartesian — Function.

cartesian(G,H) creates the Cartesian product of the two graphs. This can be abbreviated as G*H.

SimpleGraphs.char_poly — Function.

char_poly(G) returns the characteristic polynomial of adjacency(G). Use char_poly(G,function) for other possible integer matrix functions such as laplace.

SimpleGraphs.circular! — Function.

circular!(X) arranges the vertices of the graph held in the GraphEmbedding around a circle.

SimpleGraphs.circular_embedding

SimpleGraphs.code_to_tree

SimpleGraphs.complement — Function.

complement(G) creates (as a new graph) the complement of G. Note that G' is a short cut for complement(G).

SimpleGraphs.complement! — Function.

complement!(G) replaces G with its complement.

SimpleGraphs.components — Function.

components(G) returns the vertex sets of the connected components of G (as a Partition).

SimpleGraphs.contract! — Function.

contract!(G,u,v) contracts the edge (u,v) in the graph. The merged vertex is named u (hence contract!(G,v,u) results in a different, albeit isomorphic, graph).

Note: The edge (u,v) need not be present in the graph. If missing, this is equivalent to first adding the edge to the graph and then contracting it.

SimpleGraphs.cross_edges — Function.

cross_edges(G::SimpleGraph,A,B) returns the set of edges of G with one end in A and one end in B. Here A and B are collections of vertices of G.

SimpleGraphs.deg — Function.

deg(G,v) gives the degree of v in G.

deg(G) gives the degree sequence (sorted).

SimpleGraphs.deg_hist — Function.

deg_hist(G) gives a tally of how many vertices of each degree are present in the graph. Because Julia arrays are 1-based, the indexing is a bit off. Specifically, entry k in the returned array is the number of vertices of degree k-1.

SimpleGraphs.deg_sorted_vlist

Base.delete! — Function.

delete!(G,v) deletes vertex v (and any edges incident with v) from the graph.

delete!(G,v,w) deletes the edge (v,w) from G.

SimpleGraphs.diam — Function.

diam(G) returns the diameter of G or -1 if G is not connected.

SimpleGraphs.directed_ham_cycle

SimpleGraphs.disjoint_union — Function.

disjoint_union(G,H) is a new graph formed by taking disjoint copies of G and H (and no additional edges).

AbstractLattices.dist — Function.

dist(G,u,v) finds the length of a shortest path from u to v in G. Returns -1 if no such path exists.

dist(G,u) finds the distance from u to all other vertices in G. Result is returned as a Dict.

dist(G) finds all pairs of distances in G. Result is a Dict whose [u,v] entry is the distance from u to v.

SimpleGraphs.dist_matrix — Function.

dist_matrix(G) returns a matrix whose i,j-entry is the distance from the ith vertex to the jth vertex. If there is no i,j-path, that entry is -1.

SimpleGraphs.draw_labels — Function.

draw_labels(G) will add the vertices names to the drawing window.

The results are usually ugly. One can try increasing the size of the vertices (see set_vertex_size). The font size of the labels can be specified using draw_labels(G,FontSize). The default is 10.

SimpleGraphs.dual_deg — Function.

dual_deg(G,v) returns a two-tuple consisting of the out degree and in degree of the vertex v.

dual_deg(G) gives a list of all the dual degrees.

SimpleGraphs.eccentricity — Function.

eccentricity(G,v) returns the eccentricty of vertex v in the graph G. This is the maximum distance from v to another vertex (or -1 if the graph is not connected).

SimpleGraphs.edge_length — Function.

edge_length(G) returns an array containing the lengths of the edges in the current embedding of G.

edge_length(G,e) gives the distance between the embedded endpoints of the edge e in the drawing G.

LinearAlgebra.eigvals — Function.

eigvals(G) for a SimpleGraph returns the eigenvalues of G's adjacency matrix. More generally, eigvals(G,mat) returns the eigenvalues of mat(G) where mat is a matrix-valued function of G. In particular, one can use mat(G,laplace) to find the eigenvalues of G's Laplacian.

SimpleGraphs.elist — Function.

elist(G) returns the edges of G as a list (array).

SimpleGraphs.embed — Function.

embed(G) creates a new embedding for G. The full call is

embed(G,algorithm;args...)

The symbol algorithm indicates the embedding algorithm. The args are a collection of possible arguments to be sent to the algorithm.

The algorithm defaults to :circular and may be one of the following:

:circular: arrange the vertices evenly in a circle.

:random: arrange the vertices randomly.

:spring: use the spring layout from GraphLayout. Optional argument:

iterations.

:stress: use the stress layout from GraphLayout.

:spectral: use the spectral embedding. Optional arguments:

xcol – which eigenvalue to use for the x coordinate.

ycol – which eigenvalue to use for the y coordinate

Note that if the graph already has (say) an embedding, that embedding may be used as the starting point for one of the algorithms.

embed(G,d) specifies an embedding of the graph G with a dictionary d that maps vertices to coordinates (as two dimensional vectors [x,y]).

SimpleGraphs.euler — Function.

euler(G,u,v) finds an Euler trail in G starting at u and ending at v returned as a list of vertices (in the order they are visited on the trail).

euler(G,u) finds an Euler tour beginning and ending at u. Alternatively, call euler(G) and the initial/final vertex will be selected for you.

Note: The algorithm will succeed even if there are isolated vertices in the graph (just don't choose an isolated vertex as the first/last).

If no Euler trail/tour exists, an empty list is returned.

SimpleGraphs.euler_work!

SimpleGraphs.eval

SimpleGraphs.fastN! — Function.

fastN!(G,flg=true) is used to turn on (or off) fast neighborhood lookup in graphs. Switching this off decreases the size of the data structure holding the graph, but slows down look up of edges.

Note: Fast neighborhood look up is on by default.

SimpleGraphs.find_path — Function.

find_path(G,s,t) finds a shortest path from s to t. If no such path exists, an empty list is returned.

SimpleGraphs.forbid_loops! — Function.

forbid_loops!(G) disables the digraph's ability to have loops. It also removes any loops it may already have.

SimpleGraphs.geogebra — Function.

geogebra(G,file_name) takes a SimpleGraph and writes out a script to produce a drawing of this graph in GeoGebra.

Here is the secret sauce to make this work.

Run geogebra(G,file_name) to save the script to file_name. By default, the file name is geogebra.txt.

Copy the contents of file_name to the clipboard.

Create a new GeoGebra document.

In the Input zone at the bottom, enter Button[] to create a new button.

Right click the button and select the Object Properties... menu option.

Go to the Scripting tab and paste in the copied commands.

Save and close the properties window.

Press the newly created button.

Some properties of the vertices can be specified in this function with named parameters as follows:

vertex_labels: If set to true, the vertices in the drawing have labels. Default is false (no labels drawn). Note that the color of the labels matches the color of the vertices, so if you set the color to white the labels will be invisible.

vertex_color: Use this to specify the fill color of the vertices. The default is "black".

vertex_colors: This is a dictionary mapping vertices to color names. Vertices are given the color specified. If, however, a vertex is missing from this dictionary, then vertex_color is used instead.

vertex_size: Use this to specify the radius of the vertices. The default is 3.

SimpleGraphs.get_embedding_direct — Function.

get_embedding_direct(G) returns a direct reference to a graph's embedding. This function is not exposed. Users should use cache_recall(G,:GraphEmbedding) instead to retrieve a copy of the drawing. If there is not drawing, a default (circular) drawing is created.

SimpleGraphs.get_line_color — Function.

get_line_color(G) returns the color used to draw edges and the circles around vertices. See set_line_color.

SimpleGraphs.get_vertex_color — Function.

get_vertex_color(G) returns the color used to fill in vertices. See set_vertex_color.

SimpleGraphs.get_vertex_size — Function.

get_vertex_size(G) returns the size of the circle used when drawing vertices. See also set_vertex_size.

Base.getindex — Function.

Abbreviation for lex(G,H) for SimpleGraphs.

SimpleGraphs.getxy — Function.

getxy(G) returns a copy of the dictionary mapping vertices to their x,y-coordinates. This is a way of saving an embedding.

SimpleGraphs.girth — Function.

girth(G) computes the length of a shortest cycle in G or returns 0 if G is acyclic.

Warning: This implementation is quite inefficient.

SimpleGraphs.girth_cycle — Function.

girth_cycle(G) returns a shorest cycle of G as an array listing the vertices on that cycle, or an empty array if G is acyclic.

Warning: This implementation is quite inefficient.

SimpleGraphs.graffle — Function.

graffle(G::SimpleGraph, filename="julia.graffle",rad=9) creates an OmniGraffle document of this drawing.

G is the graph

filename is the name of the OmniGraffle document (be sure to end with .graffle)

rad is the radius of the circles representing the vertices

SimpleGraphs.graph_center — Function.

graph_center(G) returns the set of vertices of a SimpleGraph with minimum eccentricities.

SimpleGraphs.greedy_color — Function.

greedy_color(G,seq) creates a greedy proper coloring of the graph. The argument seq should be a 1-dimensional array containing every vertex exactly once (the function does not check this for you). The function follows that order in creating the coloring which is returned to you as a Dict mapping vertices to positive integers (representing the colors).

If seq is omitted, a random permutation of the vertices is used.

SimpleGraphs.hamCycle

SimpleGraphs.ham_check

SimpleGraphs.ham_extend

SimpleGraphs.hamiltonian_cycle — Function.

hamiltonian_cycle(G::SimpleGraph) returns a Hamiltonian cycle in the graph (if one exists) or an empty array (otherwise). This works reasonably well for small graphs.

SimpleGraphs.has — Function.

has(G,v) returns true iff v is a vertex of G.

has(G,v,w) returns true iff (v,w) is an edge of G.

SimpleGraphs.has_embedding — Function.

has_embedding(G) returns true if an embedding has been created for G in its cache.

SimpleGraphs.in_deg — Function.

in_deg(G,v) is the in degree of vertex v.

in_deg(G) is a sorted list of the in degrees of all vertices in the directed graph.

SimpleGraphs.in_neighbors — Function.

in_neighbors(G,v) gives a list of all of v's in neighbors.

SimpleGraphs.incidence — Function.

incidence(G) returns the vertex-edge incidence matrix of G.

Notes:

The result is a sparse matrix. Wrap in full to convert to nonsparse.

Each column of the matrix has exactly one +1 and one -1. If G

is undirected and an unsigned incidence matrix is desired, use incidence(G,false).

SimpleGraphs.include

SimpleGraphs.indep_poly — Function.

indep_poly(G) returns the independence polynomial of the SimpleGraph G.

SimpleGraphs.induce — Function.

induce(G,A) creates the induced subgraph of G with vertices in the set A.

SimpleGraphs.interlace — Function.

interlace(G) returns the interlace polynomial of the graph G.

SimpleGraphs.isSafe

SimpleGraphs.is_acyclic — Function.

is_acyclic(G) returns true if G has no cycles and false otherwise.

SimpleGraphs.is_connected — Function.

is_connected(G) determines if G is connected.

SimpleGraphs.is_cut_edge — Function.

is_cut_edge(G,u,v) [or is_cut_edge(G,e)] determins if (u,v) [or e] is a cut edge of G.

SimpleGraphs.is_looped — Function.

is_looped(G) indicates if the directed graph G is capable of having loops. Returning true does not mean that digraph actually has loops.

SimpleGraphs.is_strongly_connected — Function.

test if a directed graph is strongly connected by using DFS

Base.join — Function.

join(G,H) is a new graph formed by taking disjoint copies of G and H together with all possible edges between those copies.

SimpleGraphs.label_append

SimpleGraphs.laplace — Function.

laplace(G) returns the Laplacian matrix of G. This is the adjacency matrix minus the (diagonal) degree matrix. See adjacency to understand how vertices correspond to rows/columns of the resulting matrix.

SimpleGraphs.layout_spring_adj — Function.

Use the spring/repulsion model of Fruchterman and Reingold (1991):
    Attractive force:  f_a(d) =  d^2 / k
    Repulsive force:  f_r(d) = -k^2 / d
where d is distance between two vertices and the optimal distance
between vertices k is defined as C * sqrt( area / num_vertices )
where C is a parameter we can adjust

Arguments:
adj_matrix Adjacency matrix of some type. Non-zero of the eltype
           of the matrix is used to determine if a link exists,
           but currently no sense of magnitude
C          Constant to fiddle with density of resulting layout
MAXITER    Number of iterations we apply the forces
INITTEMP   Initial "temperature", controls movement per iteration

SimpleGraphs.lex — Function.

lex(G,H) computes the lexicographic product of the graphs G and H. The vertex set of this graph is the set of all ordered pairs (g,h) (with g a vertex of G and h a vertex of H in which we have the edge (g,h)~(g',h') if either g~g' in G or else g=g and h~h'.

We can use the notation G[H] also to create lex(G,H).

SimpleGraphs.line_graph — Function.

line_graph(G) creates the line graph of G.

SimpleGraphs.list2dict — Function.

list2dict(list) takes a list of (Symbol,Any) pairs and converts them to a dictionary mapping the symbols to their associated values.

SimpleGraphs.load!

SimpleGraphs.local_complement! — Function.

local_complement!(G,v) complements the edges in the neighborhood of v. That is, if u and w are neighbors of v then we toggle the edge/nonedge uw, modifying the graph.

SimpleGraphs.loops — Function.

loops(G) returns a list of vertices at which a loop is present.

SimpleGraphs.makeColorClass

SimpleGraphs.makeSimplex!

SimpleGraphs.make_scaler

SimpleGraphs.matching_poly — Function.

matching_poly(G) returns the matching polynomial of the SimpleGraph G.

SimpleGraphs.my_layout_stressmajorize_adj — Function.

Compute graph layout using stress majorization

Inputs:

δ: Matrix of pairwise distances
p: Dimension of embedding (default: 2)
w: Matrix of weights. If not specified, defaults to
       w[i,j] = δ[i,j]^-2 if δ[i,j] is nonzero, or 0 otherwise
X0: Initial guess for the layout. Coordinates are given in rows.
    If not specified, default to random matrix of Gaussians

Additional optional keyword arguments control the convergence of the algorithm and the additional output as requested:

maxiter:   Maximum number of iterations. Default: 400size(X0, 1)^2
abstols:   Absolute tolerance for convergence of stress.
           The iterations terminate if the difference between two
           successive stresses is less than abstol.
           Default: √(eps(eltype(X0))
reltols:   Relative tolerance for convergence of stress.
           The iterations terminate if the difference between two
           successive stresses relative to the current stress is less than
           reltol. Default: √(eps(eltype(X0))
abstolx:   Absolute tolerance for convergence of layout.
           The iterations terminate if the Frobenius norm of two successive
           layouts is less than abstolx. Default: √(eps(eltype(X0))
verbose:   If true, prints convergence information at each iteration.
           Default: false
returnall: If true, returns all iterates and their associated stresses.
           If false (default), returns the last iterate

Output:

The final layout X, with coordinates given in rows, unless returnall=true.

Reference:

The main equation to solve is (8) of:

@incollection{
    author = {Emden R Gansner and Yehuda Koren and Stephen North},
    title = {Graph Drawing by Stress Majorization}
    year={2005},
    isbn={978-3-540-24528-5},
    booktitle={Graph Drawing},
    seriesvolume={3383},
    series={Lecture Notes in Computer Science},
    editor={Pach, J'anos},
    doi={10.1007/978-3-540-31843-9_25},
    publisher={Springer Berlin Heidelberg},
    pages={239--250},
}

LightXML.name — Function.

name(G) returns the graph's name.

name(G,str) assigns str to be the graph's name. If str is empty, then the name is set to the default SimpleGraph{T} where T is the vertex type.

SimpleGraphs.neighbors — Function.

neighbors(G,v) returns a list of the neighbors of v.

May also be invoked as G[v].

SimpleGraphs.num_components — Function.

num_components(G) returns the number of connected components in G.

SimpleGraphs.num_trans_orientations — Function.

num_trans_orientations(G) returns the number of transitive orientations of the graph G.

SimpleGraphs.out_deg — Function.

out_deg(G,v) is the out degree of vertex v.

out_deg(G) is a sorted list of the out degrees of all vertices in the directed graph.

SimpleGraphs.out_neighbors — Function.

out_neighbors(G,v) gives a list of all v's out neighbors.

SimpleGraphs.pivot!

SimpleGraphs.private_adj

SimpleGraphs.private_dist

SimpleGraphs.radius — Function.

radius(G) returns the radius of the graph G. This is the minimum eccentricity of a vertex of G (or -1 if the graph is not connected).

SimpleGraphs.random! — Function.

random!(X) gives the graph held in X a random embedding.

SimpleGraphs.random_greedy_color — Function.

random_greedy_color(G, reps) repeatedly invokes greedy_color using random permutations of the vertex set. After reps iterations, the best coloring found is returned.

SimpleGraphs.recenter — Function.

recenter(G) translates the graph's drawing so that the center of mass of the vertices is at the origin.

SimpleGraphs.recenter!

SimpleGraphs.relabel — Function.

relabel(G) returns a copy of G in which the vertices are renamed 1:n.

relabel(G,d) (where d is a Dict) returns a copy of G in which vertex v is renamed d[v].

SimpleGraphs.remove_embedding — Function.

remove_embedding(G) erases the graph's embedding saved in its cache.

SimpleGraphs.remove_loops! — Function.

remove_loops!(G) removes all loops (if any) in the digraph, but does not alter the G's ability to have loops.

SimpleGraphs.rescale!

SimpleGraphs.rotate — Function.

rotate(G,theta) rotate's the graph's drawing by the angle theta.

SimpleGraphs.rotate!

SimpleGraphs.scale — Function.

scale(G,m) multiplies all coordinates in the graph's drawing by m. If m is omitted, the drawing is rescaled so that the average length of an edge equals 1.

SimpleGraphs.set_embedding_direct — Function.

set_embedding_direct(G,X) overwrites (or creates) the drawing X in the graph's cache. This non-exposed function should not be called by the user.

SimpleGraphs.set_fill_color

SimpleGraphs.set_line_color — Function.

set_line_color(G,color) sets the color used to draw edges and the circles around vertices.

SimpleGraphs.set_vertex_color — Function.

set_vertex_color(G,color) sets the color that gets drawn in the interior of the vertices. (All vertices get the same color.)

SimpleGraphs.set_vertex_size — Function.

set_vertex_size(G,sz) sets the size of the circle used when drawing vertices. The default is 6. See also get_vertex_size.

SimpleGraphs.show

SimpleGraphs.simplify

SimpleGraphs.spanning_forest — Function.

spanning_forest(G) creates a maximal acyclic subgraph of G.

SimpleGraphs.spectral! — Function.

spectral!(X::GraphEmbedding) gives the graph held in X an embedding based on the eigenvectors of the Laplacian matrix of the graph. Specifically, the x-coordinates come from the eigenvector associated with the second smallest eigenvalue, and the y-coordinates come from the eigenveector associated with the third smallest.

This may also be invoked as spectral!(X,xcol,ycol) to choose other eigenvectors to use for the x and y coordinates of the embedding.

SimpleGraphs.spring! — Function.

spring!(X) gives the graph held in X with a spring embedding (based on code in the GraphLayout module). If runs a default number of iterations (100) of that algorithm. To change the number of iterations, use spring!(X,nits).

SimpleGraphs.stress — Function.

Stress function to majorize

Input: X: A particular layout (coordinates in rows) d: Matrix of pairwise distances w: Weights for each pairwise distance

See (1) of Reference

SimpleGraphs.stress! — Function.

stress!(X) computes a stress major layout using code taken from the GraphLayout package.

SimpleGraphs.super_toggle! — Function.

super_toggle!(G,A,B) toggles all edges/nonedges ab where a is in A and b is in B. No error checking is done.

SimpleGraphs.tikz_file — Function.

tikz_file(G,label,filename) writes the tikz code for drawing the graph G into filename. If label is omitted (or false) vertex labels are not drawn. If filename is omitted, it defaults to graph.tex.

SimpleGraphs.tikz_print — Function.

tikz_print(G) prints tikz code to draw G. tikz_print(G,true) does likewise, with vertex labels drawn.

SimpleGraphs.tikz_string

SimpleGraphs.toggle! — Function.

toggle!(G,x,y) deletes edge xy if present or adds edge xy if absent. No error checking is done.

SimpleGraphs.transform — Function.

transform(G,A,b) applies an affine transformation to all coordinates in the graph's drawing. Here A is 2-by-2 matrix and b is a 2-vector. Each point p is mapped to A*p+b.

SimpleGraphs.transform!

SimpleGraphs.transitive_orientation — Function.

transitive_orientation(G) finds a transitive orientation of the simple graph G. The result is a SimpleDigraph. An error is raised if G does not have a transitive orientation.

SimpleGraphs.translate — Function.

translate(G,b) translates the graph's drawing by the vector b; that is, every point p in the drawing is replaced by p+b.

SimpleGraphs.translate!

SimpleGraphs.trim — Function.

trim(G) returns a copy of G with all isolated vertices removed.

trim(G,d) returns a copy of G in which we iteratively remove all vertices of degree d or smaller. For example, if G is a tree, trim(G,1) will eventually remove all vertices.

SimpleGraphs.twins — Function.

twins(G,u,v) determines if u and v are twin vertices of G. That is, if G[u]-v == G[v]-u. This is an equivalence relation.

twins(G) returns a partition of the graph's vertex set into twin equivalence classes.

SimpleGraphs.two_color — Function.

two_color(G) creates a two-coloring of the graph or throws an error if the graph is not bipartite. The output is a Dict mapping the vertex set to the values 1 and 2.

Base.union — Function.

union(G,H) creates the union of the graphs G and H. The graphs may (and typically do) have common vertices or edges.

SimpleGraphs.vertex2idx

SimpleGraphs.vertex_split — Function.

vertex_split(G) converts the directed graph G into an undirected bipartite graph. For each vertex v in G, the output graph has two vertices (v,1) and (v,2). Each edge (v,w) of G is rendered as an edge between (v,1) and (w,2) in the output graph.

SimpleGraphs.vertex_type — Function.

vertex_type(G) returns the data type of the vertices this graph may hold. For example, if G=IntGraph() then this returns Int64.`

SimpleGraphs.vlist — Function.

vlist(G) returns the vertices of G as a list (array).

SimpleGraphs.weightedlaplacian — Function.

Compute weighted Laplacian given ideal weights w

Lʷ defined in (4) of the Reference

SimpleGraphs.wiener_index — Function.

wiener_index(G) is the sum of the distances between vertices in G. Returns -1 if G is not connected.