Docstrings

Unicode_Entities.PackedNames

Unicode_Entities.Unicode_Entities — Module.

Public API (nothing is exported)

lookupname(str)

matchchar(char)

matches(str)

longestmatches(str)

completions(str)

Unicode_Entities.Unicode_Table

Unicode_Entities.VER

Unicode_Entities.__init__

Unicode_Entities._unpackword — Function.

Internal function to unpack the packed Unicode entity names

Unicode has a very limited character set for the entity names, 0-9, A-Z, -, (, and ). ( and ) are only used for 4 names, so those are special cased to store in the word table.

0 represents a space, when needed between characters 1 represents a hyphen 2-11 represents the digits 0-9 12-37 represents the letters A-Z 38-53 represents up to 16*256 words stored in the wrd2 table 54-255 represent 202 words stored in the wrd1 table

Unicode_Entities.default

Unicode_Entities.eval

Unicode_Entities.include