Docstrings

WebIO.@dom_str — Macro.

dom"div.<class>#<id>[<prop>=<value>,...]"(x...; kw...)

WebIO.@evaljs

WebIO.@js

WebIO.@js_str

WebIO.@private

WebIO.AbstractConnection

WebIO.ConnectionPool — Type.

ConnectionPool(outbox::Channel, connections::Set{AbstractConnection}=Set())

Manages the distribution of messages from the outbox channel to a set of connections. The ConnectionPool asynchronously takes messages from its outbox and sends each message to every connection in the pool. Any closed connections are automatically removed from the pool.

WebIO.Context — Type.

Scope(id; kwargs...)

An object which can send and receive messages.

Fields:

id::String: A unique ID

imports: An array of js/html/css assets to load before rendering the contents of a scope.

WebIO.DOM

WebIO.JSEvalSerialization

WebIO.JSONContext

WebIO.JSONSerialization

WebIO.JSString

WebIO.Node

WebIO.Observ

WebIO.Observable

WebIO.Scope — Type.

Scope(id; kwargs...)

An object which can send and receive messages.

Fields:

id::String: A unique ID

imports: An array of js/html/css assets to load before rendering the contents of a scope.

WebIO.Sync

WebIO.SyncCallback — Type.

A callable which updates the frontend

WebIO.WebIO

WebIO.__init__

WebIO.__inits__

WebIO._count

WebIO._dep_message_Context

WebIO._iterate

WebIO._pvec

WebIO._show

WebIO.addconnection!

WebIO.adddeps!

WebIO.append

WebIO.baseurl

WebIO.bundlejs

WebIO.camel2kebab

WebIO.children

WebIO.cssparse

WebIO.descendants_count

WebIO.devsetup

WebIO.dispatch

WebIO.encode_scripts — Function.

Any </script> tags in the js/html node representation can cause problems, because if they are printed inside a <script> tag, even if they are in quotes in a javascript string, the html parser will still read them as a closing script tag, and thus end the script content prematurely, causing untold woe.

WebIO.ensure_connection — Function.

Ensure that the pool has at least one connection, potentially blocking the current task until that is the case.

WebIO.ensure_sync

WebIO.escapeHTML

WebIO.eval

WebIO.evaljs

WebIO.handle!

WebIO.handlejs!

WebIO.htmlstring

WebIO.iframe

WebIO.import!

WebIO.include

WebIO.instanceof

WebIO.islocal

WebIO.jsexpr

WebIO.kebab2camel

WebIO.kwargs2props

WebIO.lifecycle_commands

WebIO.log

WebIO.logmsg

WebIO.lowerdeps

WebIO.makedom

WebIO.mergeprops

WebIO.mime_order — Constant.

Generic Conversion to Nodes

WebIO.newid

WebIO.node

WebIO.nodetype

WebIO.observ_id_dict

WebIO.offjs

WebIO.on

WebIO.ondependencies

WebIO.onimport

WebIO.onjs

WebIO.private

WebIO.process_messages

WebIO.promote_instanceof

WebIO.props

WebIO.providers_initialised

WebIO.re_register_renderables — Function.

Called after a provider is setup

WebIO.recmerge

WebIO.recmerge!

WebIO.register_renderable — Function.

`WebIO.register_renderable(MyType::Type)`

Registers that a WebIO.render method is available for instances of MyType. Allows WebIO to hook into the display machinery of backends such as Atom and IJulia to display the WebIO rendered version of the type as appropriate.

Also defines a Base.show(io::IO, m::MIME"text/html", x::MyType) as Base.show(io, m, WebIO.render(x))

WebIO.render — Function.

render(x::MyType)

Generic function that defines how a Julia object is rendered. Should return a Node object.

WebIO.renderable_types

WebIO.richest_html

WebIO.richest_mime

WebIO.scopes

WebIO.send_sync

WebIO.set_nosync — Function.

Set observable without synchronizing with the counterpart on the browser

WebIO.setbaseurl!

WebIO.setchild

WebIO.setchildren

WebIO.setinstanceof

WebIO.setobservable!

WebIO.setprops

WebIO.setup

WebIO.setup_comm

WebIO.setup_provider — Function.

setup_provider(s::Union{Symbol, AbstractString})

Perform any side-effects necessary to set up the given provider. For example, in IJulia, this causes the frontend to load the webio javascript bundle.

WebIO.showchildren

WebIO.showindent

WebIO.showprops

WebIO.str_interpolate

WebIO.typename

WebIO.verbose_json

WebIO.waiting_messages

WebIO.withchild

WebIO.withlastchild

WebIO.withscope