Docstrings

TiledIteration.EdgeIterator — Type.

EdgeIterator(outer, inner)

A Cartesian iterator that efficiently visits sites that are in outer but not in inner. This can be useful for calculating edge values that may require special treatment or boundary conditions.

TiledIteration.L1cachesize

TiledIteration.TileBuffer — Type.

TileBuffer(a, inds::Indices) -> v

Return a buffer-view v whose indices match inds, using the array a for storage. inds does not necessarily have to match the size of a (which allows tiles to be of different sizes, e.g., at the edges).

TileBuffer(T, inds::Indices) -> v

Return a TileBuffer, allocating a new backing array of element type T and size determined by inds.

TiledIteration.TileIterator

TiledIteration.TiledIteration

TiledIteration._getindices

TiledIteration._padded_tilesize_scale

TiledIteration._tileview

TiledIteration.cachelinesize

TiledIteration.ceildiv

TiledIteration.eval

TiledIteration.getindices

TiledIteration.include

TiledIteration.map3

TiledIteration.nextedgeitem

TiledIteration.padded_tilesize — Function.

padded_tilesize(T::Type, kernelsize::Dims, [ncache=2]) -> tilesize::Dims

Calculate a suitable tile size to approximately maximize the amount of productive work, given a stencil of size kernelsize. The element type of the array is T. Optionally specify ncache, the number of such arrays that you'd like to have fit simultaneously in L1 cache.

This favors making the first dimension larger, since the first dimension corresponds to individual cache lines.

Examples

julia> padded_tilesize(UInt8, (3,3)) (768,18)

julia> padded_tilesize(UInt8, (3,3), 4) (512,12)

julia> padded_tilesize(Float64, (3,3)) (96,18)

julia> padded_tilesize(Float32, (3,3,3)) (64,6,6)

TiledIteration.rangetype