Docstrings

JSON2.@expect

JSON2.@expectoneof

JSON2.@format — Macro.

JSON2.@format T [noargs|keywordargs] begin
    _field_ => (; options...)
    _field2_ => (; options...)
end

Specify a custom JSON formatting for a struct T. Options include: name: jsontype: omitempty: exclude: default:

By default, the JSON input is expected to contain each field of a type and be in the same order as the type was defined. For example, the struct:

struct T
    a::Int
    b::Int
    c::Union{Nothing, Int}
end

Would have JSON like:

{"a": 0, "b": 1, "c": null}
{"a": 0, "b": 1, "c": 2}
{"a": 0, "b": 1, "c": null, "d": 3} // extra fields are ignored
{"a": 0} // will work if T(a) constructor is defined
{"a": 0, "b": 1} // will work if T(a, b) constructor is defined

That is, each field must be present in the JSON input and match in position to the original struct definition. Extra arguments after the struct's own fieldtypes are ignored.

Again, the default case is for JSON input that will have consistently ordered, always-present fields; for cases where the input JSON is not well-ordered or if there is a possibility of a null, non-Union{T, Nothing} field, here's how to approach more advanced custom formatting: - If the input will always be consistenly-ordered, but fields may be missing (not null, but the field key isn't even available in the input), defaults can be provided like: JSON2.@format T begin c => (default=0,) end This says that, when reading from a JSON input, if field c isn't present, to set it's value to 0.

- If the JSON input is not consistenly-ordered, there are two other options for allowing direct type parsing:
  
  ```
  T(; a=0, b=0, c=0, kwargs...) = T(a, b, c)
  JSON2.@format T keywordargs begin
    ...
  end
  ```
  Here we've defined a "keywordargs" constructor for `T` that essentially takes a default for each field as keyword arguments, then constructs `T`.
  During parsing, the JSON input will be parsed for any valid field key-values and the keyword constructor will be called
  with whatever arguments are parsed in whatever order. Note that we also included a catchall `kwargs...` in our constructor which can be used to "throw away" or ignore any extra fields in the JSON input.

  ```
  mutable struct T
      a::Int
      b::Int
      c::Union{Nothing, Int}
  end
  T() = T(0, 0, 0)
  JSON2.@format T noargs begin
    ...
  end
  ```
  In this case, we've made `T` a _mutable_ struct and defined a "noargs" constructor `T() = ...`; we then specified in `JSON2.@format T noargs` the `noargs` option.
  During parsing, an instance of `T` will first constructed using the "noargs" constructor, then fields will be set as they're parsed from the JSON input (hence why `mutable struct` is required).

JSON2.@pretty

JSON2.BACKSLASH

JSON2.BACKSPACE

JSON2.BIG_A

JSON2.BIG_E

JSON2.BIG_F

JSON2.BIG_I

JSON2.BIG_N

JSON2.BIG_T

JSON2.BIG_Y

JSON2.BUF

JSON2.CLOSE_CURLY_BRACE

JSON2.CLOSE_SQUARE_BRACE

JSON2.COLON

JSON2.COMMA

JSON2.EMPTY_SYMBOL

JSON2.ESCAPECHARS

JSON2.FORM

JSON2.FORWARDSLASH

JSON2.Function

JSON2.JSON2

JSON2.LITTLE_A

JSON2.LITTLE_B

JSON2.LITTLE_E

JSON2.LITTLE_F

JSON2.LITTLE_I

JSON2.LITTLE_L

JSON2.LITTLE_N

JSON2.LITTLE_R

JSON2.LITTLE_S

JSON2.LITTLE_T

JSON2.LITTLE_U

JSON2.LITTLE_Y

JSON2.MINUS

JSON2.NEEDESCAPE

JSON2.NEG_ONE

JSON2.NEWLINE

JSON2.OPEN_CURLY_BRACE

JSON2.OPEN_SQUARE_BRACE

JSON2.PERIOD

JSON2.PLUS

JSON2.QUOTE

JSON2.RETURN

JSON2.SPACE

JSON2.TAB

JSON2.TEN

JSON2.ZERO

JSON2.charvalue

JSON2.escapechar

JSON2.escaped

JSON2.eval

JSON2.generate_default_read_body

JSON2.generate_missing_read_body

JSON2.generate_read_body_noargs

JSON2.generate_write_body

JSON2.getconvert

JSON2.getformats

JSON2.include

JSON2.invalid

JSON2.iscntrl

JSON2.isempty

JSON2.iswh

JSON2.needescape

JSON2.nonNothingT

JSON2.number

JSON2.peekbyte

JSON2.pretty

JSON2.read

JSON2.read_args

JSON2.readbyte

JSON2.readhexchar

JSON2.reverseescapechar

JSON2.unescape

JSON2.wh!

JSON2.write