Docstrings

Modia.@component

Modia.Instantiation.@model — Macro.

@model <Name> begin
    <declarations>
    @equations begin
        <equations>
    end
end

Fill in a Model instance with the given declarations and equations and assign it to a constant named <Name>.

Modia.BLTandPantelides — Module.

Module with graph theoretical methods for maximum matching, Pantelides index reduction and Tarjan's strongly connected components.

Author: Hilding Elmqvist, Mogram AB

Date: July-August 2016 (rewritten). Array version: Spring 2017

License: MIT

A bipartite graph (E, V) is defined by an array of arrays. Each E-entry has an integer array of indices to V-entries.

Example bipartite graph:

G = [
  [3, 5],
  [4, 6],  
  [1, 7, 9],
  [2, 8, 9],
  [1, 2]  
]

Modia.BLTandPantelidesUtilities — Module.

Module with utility functions for BLTandPantelides.

Author: Hilding Elmqvist, Mogram AB

Date: July-August 2016

License: MIT

Modia.BasicStructuralTransform — Module.

Module for structural analysis of models.

Developer: Hilding Elmqvist, Mogram AB

First version: July-August 2016

Copyright (c) 2016-2018: Hilding Elmqvist, Toivo Henningsson, Martin Otter

License: MIT (expat)

Modia.Blocks — Module.

Modia module with block component models (inspired from Modelica Standard Library).

Developer: Hilding Elmqvist, Mogram AB

Copyright (c) 2016-2018: Hilding Elmqvist, Toivo Henningsson, Martin Otter

License: MIT (expat)

Modia.Boolean — Function.

Create a boolean Variable

Modia.Date

Modia.Electric — Module.

Modia module with electric component models (inspired from Modelica Standard Library).

Developer: Hilding Elmqvist, Mogram AB

Copyright (c) 2016-2018: Hilding Elmqvist, Toivo Henningsson, Martin Otter

License: MIT (expat)

The building blocks for electric components are:

Pin - The main connector representing an electrical node with variables v and i.

OnePort - Base model for an electric device with two Pins with variables v, i, p, and n.

The following functions define variables with appropriate units:

Voltage() - Electric potential

Current() - The main flow quantity

Resistance()

Capacitance()

Modia.ExactlyRemoveSingularities — Module.

module ExactlyRemoveSingularities - exactly remove singularities from A*v = 0 with A an Integer matrix.

Use this module in the following way:

using ExactlyRemoveSingularities result = removeSingularities(A,ix,iy) # result = (iya, eqr, ix1, ix2, eqx, A1, A2) printRemoveSingularities(A,result,names)

where A is a dense or sparse Int matrix A[:,:] describing a linear system

A*v = 0

with

v[ix]: Variables used in the derivative operator, der(v[ix]), so these are the potential states.

v[iy]: Variables that must be solved from A*v = 0, because these variables do not appear in the remaining part of the model equations (e.g. absolute potentials). If variables v[iy] cannot be uniquely solved from A*v = 0, then in argument iya the variables are listed that must be set arbitrarily so that v[iy] can be uniquely solved from the equations. For example, if v[iy] contains the absolute potentials and no ground object is present, then v[iya] are the absolute potentials that must be explicitly set, e.g., by introducing an equation v[iya]=0, in order that the equations have a unique solution. Additionally, a warning message should be printed that v[iya] was set to 0.)

The function return arguments are all of type Int and have the following meaning:

iya: variables v[iya] can be arbitrarily set (iya is a sub-set of iy)

eqr: equations A[eqr,:]*v = 0 are redundant and should be removed (they do not change the result)

ix1: x[ix1] are the dependent states (ix1 is a sub-set of ix)

ix2: i[ix2] are the independent states (ix2 is a sub-set of ix)

eqx, A1, A2: equations A[eqx,:]*v = 0 should be replaced by the equations A1*x1 + A2*x2 = 0, where A1 is square, upper triangular and has full rank with A1[i,i] != 0. In other words, x1 can be computed explicitely from these equations given x2.

The calling program should do the following:

Remove equations A[eqr,:]*v = 0, because these are redundant equations.

Replace equations A[eqx,:]*v = 0 by A1*x1 + A2*x2 = 0, in order that index reduction can be performed by a structural algorithm (e.g. Pantelides algorithm).

Introduce equations v[iya] = 0 and print a warning message that these variables have been arbitrarily set to zero. This is needed, in order that the equations have a unique solution.

Modia.Execution — Module.

Modia module for executing a model including code generation and calling DAE solver.

Original developer: Toivo Henningsson, Lund

Developer: Hilding Elmqvist, Mogram AB

Copyright (c) 2016-2018: Hilding Elmqvist, Toivo Henningsson, Martin Otter

License: MIT (expat)

Modia.Float — Function.

Create a floating-point Variable

Modia.Instantiation — Module.

Modia module for instantiation and flattening of models.

Original developer: Toivo Henningsson, Lund

Developer: Hilding Elmqvist, Mogram AB

Copyright (c) 2016-2019: Hilding Elmqvist, Toivo Henningsson, Martin Otter

License: MIT (expat)

Modia.Integ — Function.

Create an integer Variable

Modia.ModelElaboration — Module.

Modia module for elaboration of models.

Developer: Hilding Elmqvist, Mogram AB

First version: July-August 2016

Copyright (c) 2016-2018: Hilding Elmqvist, Toivo Henningsson, Martin Otter

License: MIT (expat)

Modia.Modia — Module.

Modia - Dynamic Modeling and Simulation in Julia

To define a model:

  @model FirstOrder begin
     x = Variable(start=1)   # start means x(0)
     T = Parameter(0.5)      # Time constant
     u = 2.0                 # Same as Parameter(2.0)
  @equations begin
     T*der(x) + x = u        # der() means time derivative
     end
  end;

To simulate a model:

  result = simulate(FirstOrder, 2);
  @show result["x"][end];
  ModiaMath.plot(result, "x")

To run examples:

  include("/home/pfitzseb/.julia/packages/Modia/k0HTn/examples/runexamples.jl")

To run tests:

  include("/home/pfitzseb/.julia/packages/Modia/k0HTn/test/runtests.jl")

For more information, see (https://github.com/ModiaSim/Modia.jl/blob/master/README.md)

Modia.ModiaDir

Modia.ModiaLogging

Modia.Par — Function.

Shortcut for Parameter

Modia.Parameter — Function.

Create a Variable with parameter variability, meaning it is an input variable that is constant with time

Modia.Property

Modia.Rotational — Module.

Modia module with rotational component models (inspired from Modelica Standard Library).

Developer: Hilding Elmqvist, Mogram AB

Copyright (c) 2016-2018: Hilding Elmqvist, Toivo Henningsson, Martin Otter

License: MIT (expat)

Modia.StateSelection

Modia.Str — Function.

Create a string Variable

Modia.StructuralTransform — Module.

Module for structural transformation of models including alias handling and size and type deduction.

Developer: Hilding Elmqvist, Mogram AB

First version: July-August 2016

Copyright (c) 2016-2019: Hilding Elmqvist, Toivo Henningsson, Martin Otter

License: MIT (expat)

Modia.SymbolicTransform — Module.

Module for symbolic transformation of models.

Developer: Hilding Elmqvist, Mogram AB

First version: July-September 2016

License: MIT (expat)

Reference: For matrix formulas, see: Petersen, K.B, Pedersen M.S.: The Matrix Cookbook https://www.math.uwaterloo.ca/~hwolkowi/matrixcookbook.pdf

Modia.Synchronous

Modia.Utilities — Module.

Modia module with utility functions.

Developer: Hilding Elmqvist, Mogram AB

First version: July 2016

Copyright (c) 2016-2018: Hilding Elmqvist, Toivo Henningsson, Martin Otter

License: MIT (expat)

Modia.Var — Function.

Shortcut for Variable

Modia.Instantiation.Variability — Type.

Enum for variability of a variable

source

Modia.Instantiation.Variable — Type.

A constructor for a Variable, the main object for a model variable with attributes

Keyword arguments

value = undefined: the value of the Variable

info = "": documentation string

unit: unit of measure: defaults to no units unless provided with the value

displayUnit = unit: unit used for display

min = undefined: minimum value

max = undefined: maximum value

start = undefined: starting value

fixed::Bool = false: fixed value

nominal = undefined: nominal value

variability = continuous: other options include parameter, constant, or discrete

T: type of the value; taken from value

size::Tuple = (): size of the variable

flow::Bool = false: indicates a flow variable for connectors

state::Bool = true: indicates a state Variable

property = general: other options include symmetric, orthogonal, and rotationGroup3D

Modia.Version

Modia.addComponent!

Modia.allInstances

Modia.ModelElaboration.checkSimulation — Function.

function checkSimulation(mod, stopTime, observer, finalSolution; options...)

Simulates model mod until stopTime and checks that the final value of the observer variable is approximately finalSolution. The following options are available:

logTranslation: determines if logging of the translation is performed

...

Modia.Instantiation.constant — Constant.

Indicates a constant valued variable

Modia.Instantiation.continuous — Constant.

Indicates a continuous variable

Modia.Instantiation.discrete — Constant.

Indicates a discrete variable

Modia.eval

Modia.general

Modia.include

Modia.orthogonal

Modia.Instantiation.parameter — Constant.

Indicates an input value that stays constant with time

Modia.rotationGroup3D

Modia.simulate

Modia.simulateModel

Modia.ModelElaboration.simulateMultiModeModel — Function.

Experimental code for multi-mode handling with impulses.

Modia.skew

Modia.skewCoords

Modia.symmetric

Modia.undefined — Constant.

A value meant to be filled in later