Docstrings

Yao.Registers.@bit_str — Macro.

@bit_str -> QuBitStr

Construct a bit string. such as bit"0000". The bit strings also supports string concat. Just use it like normal strings.

Yao.Blocks.ConstGateTools.@const_gate — Macro.

@const_gate NAME = MAT_EXPR
@const_gate NAME::Type = MAT_EXPR
@const_Gate NAME::Type

This macro simplify the definition of a constant gate. It will automatically bind the matrix form to a constant which will reduce memory allocation in the runtime.

@const_gate X = ComplexF64[0 1;1 0]

or

@const_gate X::ComplexF64 = [0 1;1 0]

You can bind new element types by simply re-declare with a type annotation.

@const_gate X::ComplexF32

Yao.Interfaces.@fn — Macro.

macro fn([name,] f)

Define a in-place function on a register inside circuits.

Yao.Registers.AbstractRegister — Type.

AbstractRegister{B, T}

abstract type that registers will subtype from. B is the batch size, T is the data type.

Required Properties

Property	Description	default

nqubits(reg)	get the total number of qubits.

nactive(reg)	get the number of active qubits.

nremain(reg)	get the number of remained qubits.	nqubits - nactive

nbatch(reg)	get the number of batch.	B

state(reg)	get the state of this register. It always return the matrix stored inside.

statevec(reg)	get the raveled state of this register. .

hypercubic(reg)	get the hypercubic form of this register. .

eltype(reg)	get the element type stored by this register on classical memory. (the type Julia should use to represent amplitude)	T

copy(reg)	copy this register.

similar(reg)	construct a new register with similar configuration.

Required Methods

Multiply

*(op, reg)

define how operator op act on this register. This is quite useful when there is a special approach to apply an operator on this register. (e.g a register with no batch, or a register with a MPS state, etc.)

Note

be careful, generally, operators can only be applied to a register, thus we should only overload this operation and do not overload *(reg, op).

Pack Address

pack addrs together to the first k-dimensions.

Example

Given a register with dimension [2, 3, 1, 5, 4], we pack [5, 4] to the first 2 dimensions. We will get [5, 4, 2, 3, 1].

Focus Address

focus!(reg, range)

merge address in range together as one dimension (the active space).

Example

Given a register with dimension (2^4)x3 and address [1, 2, 3, 4], we focus address [3, 4], will pack [3, 4] together and merge them as the active space. Then we will have a register with size 2^2x(2^2x3), and address [3, 4, 1, 2].

Initializers

Initializers are functions that provide specific quantum states, e.g zero states, random states, GHZ states and etc.

register(::Type{RT}, raw, nbatch)

an general initializer for input raw state array.

register(::Val{InitMethod}, ::Type{RT}, ::Type{T}, n, nbatch)

init register type RT with InitMethod type (e.g Val{:zero}) with element type T and total number qubits n with nbatch. This will be auto-binded to some shortcuts like zero_state, rand_state.

Yao.Blocks

Yao.Boost

Yao.C

Yao.Registers.DefaultRegister — Type.

DefaultRegister{B, T} <: AbstractRegister{B, T}

Default type for a quantum register. It contains a dense array that represents a batched quantum state with batch size B of type T.

Yao.DefaultType

Yao.DensityMatrix

Yao.ENVNAME

Yao.Blocks.H — Constant.

H

The Hadamard gate acts on a single qubit. It maps the basis state $|0\rangle$ to $\frac{|0\rangle + |1\rangle}{\sqrt{2}}$ and $|1\rangle$ to $\frac{|0\rangle - |1\rangle}{\sqrt{2}}$, which means that a measurement will have equal probabilities to become 1 or 0. It is representated by the Hadamard matrix:

\[H = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1 \\ 1 & -1 \end{pmatrix}\]

Yao.I2

Yao.Interfaces

Yao.Intrinsics

Yao.Interfaces.InvOrder — Constant.

InvOrder

Return a FunctionBlock of inversing the order.

Yao.MEASURE

Yao.MEASURE_REMOVE

Yao.PKGNAME

Yao.Register

Yao.Registers

Yao.Interfaces.Reset — Constant.

Reset

Yao.Interfaces.Rx — Function.

Rx([type=Yao.DefaultType], theta) -> RotationGate{1, type, X}

Returns a rotation X gate.

Yao.Interfaces.Ry — Function.

Ry([type=Yao.DefaultType], theta) -> RotationGate{1, type, Y}

Returns a rotation Y gate.

Yao.Interfaces.Rz — Function.

Rz([type=Yao.DefaultType], theta) -> RotationGate{1, type, Z}

Returns a rotation Z gate.

Yao.Signal

Yao.Blocks.X — Constant.

X

The Pauli-X gate acts on a single qubit. It is the quantum equivalent of the NOT gate for classical computers (with respect to the standard basis $|0\rangle$, $|1\rangle$). It is represented by the Pauli X matrix:

\[X = \begin{pmatrix} 0 & 1\\ 1 & 0 \end{pmatrix}\]

Yao.Blocks.Y — Constant.

Y

The Pauli-Y gate acts on a single qubit. It equates to a rotation around the Y-axis of the Bloch sphere by $\pi$ radians. It maps $|0\rangle$ to $i|1\rangle$ and $|1\rangle$ to $-i|0\rangle$. It is represented by the Pauli Y matrix:

\[Y = \begin{pmatrix} 0 & -i\\ i & 0 \end{pmatrix}\]

Yao.Yao — Module.

Extensible Framework for Quantum Algorithm Design for Humans.

简单易用可扩展的量子算法设计框架。

Yao.Blocks.Z — Constant.

Z

The Pauli-Z gate acts on a single qubit. It equates to a rotation around the Z-axis of the Bloch sphere by $\pi$ radians. Thus, it is a special case of a phase shift gate (see shift) with $\theta = \pi$. It leaves the basis state $|0\rangle$ unchanged and maps $|1\rangle$ to $-|1\rangle$. Due to this nature, it is sometimes called phase-flip. It is represented by the Pauli Z matrix:

\[Z = \begin{pmatrix} 1 & 0\\ 0 & -1 \end{pmatrix}\]

Yao.Interfaces.addbit — Function.

addbit(n::Int) -> FunctionBlock{:AddBit}

Return a FunctionBlock of adding n bits.

Yao.addbit!

Yao.adjoint

Yao.Blocks.apply! — Function.

apply!(reg, block, [signal])

apply a block to a register reg with or without a cache signal.

Yao.asindex

Yao.Intrinsics.basis — Function.

basis(num_bit::Int) -> UnitRange{Int}
basis(state::AbstractArray) -> UnitRange{Int}

Returns the UnitRange for basis in Hilbert Space of num_bit qubits. If an array is supplied, it will return a basis having the same size with the first diemension of array.

Yao.Blocks.blocks — Function.

blocks(composite_block)

get an iterator that iterate through all sub-blocks.

Yao.cache

Yao.Interfaces.chain — Function.

chain([T], n::Int) -> ChainBlock
chain([n], blocks) -> ChainBlock

Returns a ChainBlock. This factory method can be called lazily if you missed the total number of qubits.

This chains several blocks with the same size together.

Yao.clearall!

Yao.Interfaces.concentrate — Function.

concentrate(nbit::Int, block::AbstractBlock, addrs) -> Concentrator{nbit}

concentrate blocks on serveral addrs.

Yao.Interfaces.control — Function.

control([total], controls, target) -> ControlBlock

Constructs a ControlBlock

Yao.Blocks.datatype — Function.

datatype(x) -> DataType

Returns the data type of x.

Yao.Registers.density_matrix — Function.

density_matrix(register)

Returns the density matrix of this register.

Yao.Blocks.dispatch! — Function.

dispatch!(block, params)
dispatch!(block, params...)

dispatch parameters to this block.

Yao.eval

Yao.Registers.extend! — Function.

extend!(r::DefaultRegister, n::Int) -> DefaultRegister
extend!(n::Int) -> Function

extend the register by n bits in state |0>. i.e. |psi> -> |000> ⊗ |psi>, extended bits have higher indices. If only an integer is provided, then perform lazy evaluation.

Yao.Registers.fidelity — Function.

fidelity(reg1::DefaultRegister, reg2::DefaultRegister) -> Vector

Yao.focus

Yao.Registers.focus! — Function.

focus!(reg::DefaultRegister, bits::Ints) -> DefaultRegister
focus!(locs::Int...) -> Function

Focus register on specified active bits.

Yao.Intrinsics.hypercubic — Function.

hypercubic(r::AbstractRegister) -> AbstractArray

Return the hypercubic form (high dimensional tensor) of this register, only active qubits are considered.

Yao.include

Yao.invorder

Yao.invorder!

Yao.iscacheable

Yao.iscached

Yao.Registers.isnormalized — Function.

isnormalized(reg::DefaultRegister) -> Bool

Return true if a register is normalized else false.

Yao.Blocks.mat — Function.

mat(block) -> Matrix

Returns the matrix form of this block.

Yao.Interfaces.matrixgate — Function.

matrixgate(matrix::AbstractMatrix) -> GeneralMatrixGate
matrixgate(matrix::MatrixBlock) -> GeneralMatrixGate

Construct a general matrix gate.

Yao.Registers.measure — Function.

measure(register, [n=1]) -> Vector

measure active qubits for n times.

Yao.measure!

Yao.Registers.measure_remove! — Function.

measure_remove!(register)

measure the active qubits of this register and remove them.

Yao.nactive — Function.

nactive(x) -> Int

Returns number of active qubits

Yao.nbatch

Yao.Blocks.nparameters — Function.

nparameters(x) -> Integer

Returns the number of parameters of x.

Yao.nqubits — Function.

nqubits(m::AbstractRegister) -> Int

Returns number of qubits in a register,

nqubits(m::AbstractBlock) -> Int

Returns number of qubits applied for a block,

nqubits(m::AbstractArray) -> Int

Returns size of the first dimension of an array, in 2^nqubits.

Yao.nremain

Yao.Blocks.parameters — Function.

parameters(block) -> Vector

Returns a list of all parameters in block.

Yao.Interfaces.phase — Function.

phase([type=Yao.DefaultType], theta) -> PhaseGate{:global}

Returns a global phase gate.

Yao.probs

Yao.Registers.product_state — Function.

product_state(::Type{T}, n::Int, config::Int, nbatch::Int=1) -> DefaultRegister

a product state on given configuration config, e.g. product_state(ComplexF64, 5, 0) will give a zero state on a 5 qubit register.

Yao.pull

Yao.Interfaces.put — Function.

put([total::Int, ]pa::Pair) -> PutBlock{total}

put a block at the specific position(s), can be lazy constructed.

Yao.Registers.rand_state — Function.

rand_state(::Type{T}, n::Int, nbatch::Int=1) -> DefaultRegister

here, random complex numbers are generated using randn(ComplexF64).

Yao.rank3

Yao.Interfaces.reflect — Function.

reflect(mirror::DefaultRegister{1}) -> ReflectBlock
reflect(mirror::Vector) -> ReflectBlock

Return an ReflectBlock along with state vector mirror as the axis.

Yao.Registers.register — Function.

register([type], bit_str, [nbatch=1]) -> DefaultRegister

Returns a DefaultRegister by inputing a bit string, e.g

using Yao
register(bit"0000")

register(raw) -> DefaultRegister

Returns a DefaultRegister from a raw dense array (Vector or Matrix).

Yao.Registers.relax! — Function.

relax!(reg::DefaultRegister; nbit::Int=nqubits(reg)) -> DefaultRegister
relax!(reg::DefaultRegister, bits::Ints; nbit::Int=nqubits(reg)) -> DefaultRegister
relax!(bits::Ints...; nbit::Int=-1) -> Function

Inverse transformation of focus, with nbit is the number of active bits of target register.

Yao.reorder — Function.

Reorder the lines of qubits.

Yao.reorder!

Yao.reset!

Yao.Interfaces.roll — Function.

roll([n::Int, ], blocks...,) -> Roller{n}

Construct a Roller block, which is a faster than KronBlock to calculate similar small blocks tile on the whole address.

Yao.Interfaces.rollrepeat — Function.

rollrepeat([n::Int,] block::MatrixBlock) -> Roller{n}

Construct a Roller block, which is a faster than KronBlock to calculate similar small blocks tile on the whole address.

Yao.Interfaces.rot — Function.

rot([type=Yao.DefaultType], U, theta) -> RotationGate{N, type, U}

Returns an arbitrary rotation gate on U.

Yao.Registers.select — Function.

select(reg::AbstractRegister, b::Integer) -> AbstractRegister

the non-inplace version of select! function.

Yao.Registers.select! — Function.

select!(reg::AbstractRegister, b::Integer) -> AbstractRegister
select!(b::Integer) -> Function

select specific component of qubit, the inplace version, the currified version will return a Function.

e.g. select!(reg, 0b110) will select the subspace with (focused) configuration 110. After selection, the focused qubit space is 0, so you may want call relax! manually.

Yao.Interfaces.sequence — Function.

Returns a Sequential block. This factory method can be called lazily if you missed the total number of qubits.

This is the loose version of sequence, that does not support the mat related interfaces.

Yao.Interfaces.shift — Function.

shift([type=Yao.DefaultType], theta) -> PhaseGate{:shift}

Returns a phase shift gate.

Yao.signal

Yao.Registers.stack — Function.

stack(regs::DefaultRegister...) -> DefaultRegister

stack multiple registers into a batch.

Yao.state

Yao.Registers.statevec — Function.

statevec(r::AbstractRegister) -> AbstractArray

Return the raveled state (vector) form of this register.

Yao.Interfaces.swap — Function.

swap([n], [type], line1, line2) -> Swap

Returns a swap gate on line1 and line2

Yao.Registers.tracedist — Function.

tracedist(reg1::DefaultRegister, reg2::DefaultRegister) -> Vector
tracedist(reg1::DensityMatrix, reg2::DensityMatrix) -> Vector

trace distance.

Yao.Registers.uniform_state — Function.

uniform_state(::Type{T}, n::Int, nbatch::Int=1) -> DefaultRegister

uniform state, the state after applying H gates on |0> state.

Yao.update_cache

Yao.Registers.zero_state — Function.

zero_state(::Type{T}, n::Int, nbatch::Int=1) -> DefaultRegister

Yao.Registers.ρ — Function.

ρ(register)

Returns the density matrix of this register.

Yao.⊗

Yao.幺 — Module.

Extensible Framework for Quantum Algorithm Design for Humans.

简单易用可扩展的量子算法设计框架。

幺 means unitary in Chinese.