Docstrings

SteamTables.B23 — Function.

B23(InputType::Symbol, InputValue)

Returns the boundary between regions 2 and 3. InputType is either :T or :P to indicate that InputValue is temperature [K] or pressure [MPa]. The complimentary value is returned.

SteamTables.B2bc — Function.

B2bc

Returns the boundary between regions 2b and 2c, approx s=5.85kJ/kgK
InputType is either :h or :P to indicate that InputValue is enthalpy [kJ/kg]
or pressure [MPa]. The complimentary value is returned.
Valid from saturation line at 554.485K & 6.54670MPa to 1019.32K & 100MPa

SteamTables.Mr

SteamTables.P3

SteamTables.Pc

SteamTables.Psat — Function.

Psat

Utility function that returns the vapour pressure at temperature T [K].

SteamTables.R

SteamTables.Region1 — Function.

Region1

Returns all the property values in region 1.
273.15K ≤ T ≤ 623.15K  Psat(T) ≤ P ≤ 100MPa
Pressures in MPa and temperature in [K]
The property to be returned is specified in Output.
    :SpecificG        kJ/kg
    :SpecificF        kJ/kg
    :SpecificV        m3/kg
    :SpecificU        kJ/kg
    :SpecificS        kJ/kgK
    :SpecificH        kJ/kg
    :SpecificCP       kJ/kgK
    :SpecificCV       kJ/kgK
    :SpeedOfSound     m/s

SteamTables.Region1_TPh — Function.

Region1_TPh

Returns T [K] from P[MPa] and h[kJ/kg] in Region 1.
273.15K ≤ T ≤ 623.15K  Psat(T) ≤ P ≤ 100MPa
The inconsistency with the main Region 1 model is ~25mK.
If you need better accuracy (why???), start with this value and iterate.

SteamTables.Region1_TPs — Function.

Region1_TPs

Returns T [K] from P[MPa] and s[kJ/kgK] in Region 1.
273.15K ≤ T ≤ 623.15K  Psat(T) ≤ P ≤ 100MPa
The inconsistency with the main Region 1 model is ~25mK.
If you need better accuracy (why???), start with this value and iterate.

SteamTables.Region2 — Function.

Region2

Returns all the property values in region 2.
273.15K ≤ T ≤ 623.15K  0 ≤ P ≤ Psat(T)
623.15K <  T ≤ 863.15K  0 <  P ≤ P(T) from B23-model
863.15K <  T ≤ 1073.15K 0 <  P ≤ 100MPa
Accuracy in the metastable region is reasonable above 10MPa, only
Pressures in MPa and temperature in [K]
The property to be returned is specified in Output.
    :SpecificG        kJ/kg
    :SpecificF        kJ/kg        
    :SpecificV        m3/kg
    :SpecificU        kJ/kg
    :SpecificS        kJ/kgK
    :SpecificH        kJ/kg
    :SpecificCP       kJ/kgK
    :SpecificCV       kJ/kgK
    :SpeedOfSound     m/s

SteamTables.Region2a_TPh — Function.

Region2a_TPh

Returns T [K] from P[MPa] and h[kJ/kg] in Region 2a.
Pressures in MPa and temperature in [K]

SteamTables.Region2a_TPs — Function.

Region2a_TPs

Returns T [K] from P[MPa] and s[kJ/kgK] in Region 2a.
Pressures in MPa and temperature in [K]

SteamTables.Region2b_TPh — Function.

Region2b_TPh

Returns T [K] from P[MPa] and h[kJ/kg] in Region 2b.
Pressures in MPa and temperature in [K]

SteamTables.Region2b_TPs — Function.

Region2b_TPs

Returns T [K] from P[MPa] and s[kJ/kgK] in Region 2a.
Pressures in MPa and temperature in [K]

SteamTables.Region2c_TPh — Function.

Region2c_TPh

Returns T [K] from P[MPa] and h[kJ/kg] in Region 2b.
Pressures in MPa and temperature in [K]

SteamTables.Region2c_TPs — Function.

Region2c_TPs

Returns T [K] from P[MPa] and s[kJ/kgK] in Region 2a.
Pressures in MPa and temperature in [K]

SteamTables.Region2meta — Function.

Region2meta

Returns all the property values in region 2 in the metastable region,
from the saturated-vapour line to the 5% equlibrium moisture line, a.k.a the
practical Wilson line:
    %equilibrium moisture = [h - h_liq(P)] / [h_vap(P) - h_liq(P)]
for pressures from P3 up to 10MPa.
Pressures in MPa and temperature in [K]
The property to be returned is specified in Output.
    :SpecificG        kJ/kg       
    :SpecificF        kJ/kg  
    :SpecificV        m3/kg
    :SpecificU        kJ/kg
    :SpecificS        kJ/kgK
    :SpecificH        kJ/kg
    :SpecificCP       kJ/kgK
    :SpecificCV       kJ/kgK
    :SpeedOfSound     m/s

SteamTables.Region3 — Function.

Initialise from ideal gas law, then use root finder to calculate P by iterating on Region3ρ.
Pass through properties from Region3ρ

SteamTables.Region3_ρ — Function.

Region3_ρ

Returns all the property values in region 3.
623.15K ≤ T ≤ T(P) from B23-model P(T) from B23-model ≤ P ≤ 100MPa
Pressures in MPa and temperature in [K]
The property to be returned is specified in Output.
    :SpecificG        kJ/kg    
    :SpecificF        kJ/kg
    :Pressure         MPa
    :SpecificU        kJ/kg
    :SpecificS        kJ/kgK
    :SpecificH        kJ/kg
    :SpecificCP       kJ/kgK
    :SpecificCV       kJ/kgK
    :SpeedOfSound     m/s

SteamTables.Region4 — Function.

Region4

Returns the vapour-liquid phase boundary (saturation line).
Valid from triple point to critical point
273.15K ≤ T ≤ 647.096K
InputType is either :T or :P to indicate that InputValue is temperature [K]
or pressure [MPa]. The complimentary value is returned.

SteamTables.Region5 — Function.

Region5

Returns all the property values in region 5.
1073.15K ≤ T ≤ 2273.15K  0 ≤ P ≤ 50MPa
Pressures in MPa and temperature in [K]

The property to be returned is specified in Output.
    :SpecificG        kJ/kg
    :SpecificF        kJ/kg
    :SpecificV        m3/kg
    :SpecificU        kJ/kg
    :SpecificS        kJ/kgK
    :SpecificH        kJ/kg
    :SpecificCP       kJ/kgK
    :SpecificCV       kJ/kgK
    :SpeedOfSound     m/s

SteamTables.RegionID

SteamTables.RegionID_Ph — Function.

RegionID_Ph

Identifies the applicable region, based on specified P [MPa] and h [kJ/kg].
This allows the correct function to be called to retrieve properties.
Backwards equations are only available for Regions 1 and 2 and the vapour-liquid
equilibrium line, which is only used when explicitly called.

As there are no explicit domains defined for (T, h), the method is iterative.
The reverse equation for T = f(P,h) is called for each region and the result
is checked against the bounds for the region.

SteamTables.RegionID_Ps — Function.

RegionID_Ps

Identifies the applicable region, based on specified P [MPa] and h [kJ/kg].
This allows the correct function to be called to retrieve properties.
Backwards equations are only available for Regions 1 and 2 and the vapour-liquid
equilibrium line, which is only used when explicitly called.

As there are no explicit domains defined for (T, s), the method is iterative.
The reverse equation for T = f(P,s) is called for each region and the result
is checked against the bounds for the region.

SteamTables.SpecificCP — Function.

SpecificCP

Utility function that returns the specific isobaric heat capacity [kJ/kgK] from P [MPa] and T [K]

SteamTables.SpecificCP_Ph — Function.

SpecificCP_Ph

Utility function that returns the specific entropy [kJ/kgK] from P [MPa] and h [kJ/kg]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificCP_Ps — Function.

SpecificCP_Ps

Utility function that returns the speciic entropy [kJ/kgK] from P [MPa] and s [kJ/kgK]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificCV — Function.

SpecificCV

Utility function that returns the specific isochoric heat capacity [kJ/kgK] from P [MPa] and T [K]

SteamTables.SpecificCV_Ph — Function.

SpecificCV_Ph

Utility function that returns the specific entropy [kJ/kgK] from P [MPa] and h [kJ/kg]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificCV_Ps — Function.

SpecificCV_Ps

Utility function that returns the speciic entropy [kJ/kgK] from P [MPa] and s [kJ/kgK]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificF — Function.

SpecificF

Utility function that returns the Helmholtz free energy [kJ/kgK] from P [MPa] and T [K]

SteamTables.SpecificF_Ph — Function.

SpecificF_Ph

Utility function that returns the Helmholtz free energy [kJ/kgK] from P [MPa] and h [kJ/kg]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificF_Ps — Function.

SpecificF_Ps

Utility function that returns the Helmholtz free energy [kJ/kgK] from P [MPa] and s [kJ/kgK]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificG — Function.

SpecificG

Utility function that returns the Gibbs free energy [kJ/kgK] from P [MPa] and T [K]

SteamTables.SpecificG_Ph — Function.

SpecificG_Ph

Utility function that returns the Gibbs free energy [kJ/kgK] from P [MPa] and h [kJ/kg]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificG_Ps — Function.

SpecificG_Ps

Utility function that returns the Gibbs free energy [kJ/kgK] from P [MPa] and s [kJ/kgK]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificH — Function.

SpecificH

Utility function that returns the specific enthalpy [kJ/kg] from P [MPa] and T [K]

SteamTables.SpecificH_Ps — Function.

SpecificH_Ps

Utility function that returns the speciic enthalpy [kJ/kg] from P [MPa] and s [kJ/kgK]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificS — Function.

SpecificS

Utility function that returns the specific entropy [kJ/kgK] from P [MPa] and T [K]

SteamTables.SpecificS_Ph — Function.

SpecificS_Ph

Utility function that returns the specific entropy [kJ/kgK] from P [MPa] and h [kJ/kg]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificU — Function.

SpecificU

Utility function that returns the specific internal energy [kJ/kg] from P [MPa] and T [K]

SteamTables.SpecificU_Ph — Function.

SpecificU_Ph

Utility function that returns the internal energy [kJ/kgK] from P [MPa] and h [kJ/kg]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificU_Ps — Function.

SpecificU_Ps

Utility function that returns the internal energy [kJ/kgK] from P [MPa] and s [kJ/kgK]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificV — Function.

SpecificV

Utility function that returns the specific volume [m3/kg] from P [MPa] and T [K]

SteamTables.SpecificV_Ph — Function.

SpecificV_Ph

Utility function that returns the specific volume [m3/kg] from P [MPa] and h [kJ/kg]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpecificV_Ps — Function.

SpecificV_Ps

Utility function that returns the specific volume [m3/kg] from P [MPa] and s [kJ/kgK]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpeedOfSound — Function.

SpeedOfSound

Utility function that returns the sonic velocity [m/s] from P [MPa] and T [K]

SteamTables.SpeedOfSound_Ph — Function.

SpeedOfSound_Ph

Utility function that returns the specific entropy [kJ/kgK] from P [MPa] and h [kJ/kg]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SpeedOfSound_Ps — Function.

SpeedOfSound_Ps

Utility function that returns the speciic entropy [kJ/kgK] from P [MPa] and s [kJ/kgK]
The explicit backwards equations are only available in regions 1 and 2. Input outside these
will result in a DomainError exception.

SteamTables.SteamTables — Module.

SteamTables

IAPWS-97 Industrial Formulation Properties of water and steam. Provides the Gibbs and Helmholtz free energies, enthalpy, entropy Cp, Cv and sonic velocity given inputs that are eith P&T, P&h or P&s.

Exported functions:

Single input:

Psat(T) and Tsat(P) returns the saturation linearindices

Two inputs:

P and T SpecificG, SpecificF, SpecificV, SpecificU, SpecificS, SpecificH, SpecificCP, SpecificCV, SpeedOfSound

P and h SpecificGPh, SpecificFPh, SpecificVPh, SpecificUPh, SpecificSPh, SpecificHPh, SpecificCPPh, SpecificCVPh, SpeedOfSound_Ph

P and s SpecificGPs, SpecificFPs, SpecificVPs, SpecificUPs, SpecificSPs, SpecificHPs, SpecificCPPs, SpecificCVPs, SpeedOfSound_Ps

SpecificG [kJ/kg] Specific Gibbs free energy SpecificF [kJ/kg] Specific Helmholtz free energy SpecificV [m3/kg] Specific volume SpecificU [kJ/kg] Specific internal energy SpecificS [kJ/kgK] Specific entropy SpecificH [kJ/kg] Specific enthalpy SpecificCp [kJ/kgK] Specific isobaric heat capacity SpecificCv [kJ/kgK] Specific isochoric heat capacity SpeedOfSound [m/s] Sonic velocity

Temperatures in K, Pressures in MPa

##Exported constants R = 0.461526 [kJ/kg/K] Universal gas constant Tc = 647.096 [K] Critical temperature of water Pc = 22.064 [kg/m3] Critical density of water T3 = 273.16 [K] Triple point temperature of water P3 = 611.657E-6 [MPa] Triple point pressure of water Mr = 18.01528 [kg/kmol] Molecular weight of water

SteamTables.T3

SteamTables.Tc

SteamTables.Tsat — Function.

Tsat

Utility function that returns the vapour pressure at pressure P [MPa]

SteamTables.eval

SteamTables.include

SteamTables.ρc