Docstrings

ReplMaker.ReplMaker

ReplMaker.eval

ReplMaker.include

ReplMaker.initrepl — Function.

         initrepl(parser::Function;
                  prompt_text = "myrepl> ",
                  prompt_color = :blue,
                  start_key = ')',
                  repl = Base.active_repl,
                  mode_name = :mylang,
                  valid_input_checker::Function = (s -> true),
                  keymap::Dict = REPL.LineEdit.default_keymap_dict,
                  completion_provider = REPL.REPLCompletionProvider(),
                  sticky = true,
                  startup_text=true
                  )

creates a custom repl mode which takes in code and parses it according to whatever parsing function you provide in the argument parser. Choose which key initializes the repl mode with start_key, the name of your repl mode with mode_name and optionally provide a function which checks if a given repl input is valid before parsing with valid_input_checker. Autocompletion options are supplied through the argument completion_provider which defaults to the standard julia REPL TAB completions.