Docstrings

ReinforcementLearningEnvironmentDiscrete.AbsorbingDetMDP — Function.

AbsorbingDetMDP(;ns = 10^3, na = 10)

Returns a random deterministic absorbing MDP

ReinforcementLearningEnvironmentDiscrete.CliffWalking

ReinforcementLearningEnvironmentDiscrete.DetMDP — Function.

DetMDP(; ns = 10^4, na = 10)

Returns a random deterministic MDP.

ReinforcementLearningEnvironmentDiscrete.DetTreeMDP — Function.

DetTreeMDP(; na = 4, depth = 5)

Returns a treeMDP with random rewards at the leaf nodes.

ReinforcementLearningEnvironmentDiscrete.DetTreeMDPwithinrew — Function.

DetTreeMDPwithinrew(; na = 4, depth = 5)

Returns a treeMDP with random rewards.

ReinforcementLearningEnvironmentDiscrete.DiscreteMaze — Type.

struct DiscreteMaze
    mdp::MDP
    maze::Array{Int64, 2}
    goals::Array{Int64, 1}
    nzpos::Array{Int64, 1}

ReinforcementLearningEnvironmentDiscrete.MDP — Type.

mutable struct MDP 
    ns::Int64
    na::Int64
    state::Int64
    trans_probs::Array{AbstractArray, 2}
    reward::Array{Float64, 2}
    initialstates::Array{Int64, 1}
    isterminal::Array{Int64, 1}

A Markov Decision Process with ns states, na actions, current state, naxns - array of transition probabilites trans_props which consists for every (action, state) pair of a (potentially sparse) array that sums to 1 (see getprobvecrandom, getprobvecuniform, getprobvecdeterministic for helpers to constract the transition probabilities) naxns - array of reward, array of initial states initialstates, and ns - array of 0/1 indicating if a state is terminal.

ReinforcementLearningEnvironmentDiscrete.MDPEnv

ReinforcementLearningEnvironmentDiscrete.POMDPEnv

ReinforcementLearningEnvironmentDiscrete.ReinforcementLearningEnvironmentDiscrete

ReinforcementLearningEnvironmentDiscrete.StochMDP — Function.

StochMDP(; na = 10, ns = 50) = MDP(ns, na)

ReinforcementLearningEnvironmentDiscrete.StochTreeMDP — Function.

StochTreeMDP(; na = 4, depth = 4, bf = 2)

Returns a random stochastic treeMDP with branching factor bf.

ReinforcementLearningEnvironmentDiscrete.addrandomwall!

ReinforcementLearningEnvironmentDiscrete.breaksomewalls

ReinforcementLearningEnvironmentDiscrete.checkpos

ReinforcementLearningEnvironmentDiscrete.emptytransprob!

ReinforcementLearningEnvironmentDiscrete.eval

ReinforcementLearningEnvironmentDiscrete.getemptymaze

ReinforcementLearningEnvironmentDiscrete.getprobvecdeterministic — Function.

getprobvecdeterministic(n, min = 1, max = n)

Returns a SparseVector of length n where one element in min:max has value 1.

ReinforcementLearningEnvironmentDiscrete.getprobvecrandom — Function.

getprobvecrandom(n)

Returns an array of length n that sums to 1. More precisely, the array is a sample of a Dirichlet distribution with n categories and $α_1 = ⋯ = α_n = 1$.

getprobvecrandom(n, min, max)

Returns an array of length n that sums to 1 where all elements outside of min:max are zero.

ReinforcementLearningEnvironmentDiscrete.getprobvecuniform — Function.

getprobvecuniform(n)  = fill(1/n, n)

ReinforcementLearningEnvironmentDiscrete.include

ReinforcementLearningEnvironmentDiscrete.indto2d

ReinforcementLearningEnvironmentDiscrete.mazetomdp

ReinforcementLearningEnvironmentDiscrete.observationindex

ReinforcementLearningEnvironmentDiscrete.posto1d

ReinforcementLearningEnvironmentDiscrete.rng

ReinforcementLearningEnvironmentDiscrete.run! — Function.

run!(mdp::MDP, action::Int64)

Transition to a new state given action. Returns the new state.

run!(mdp::MDP, policy::Array{Int64, 1}) = run!(mdp, policy[mdp.state])

ReinforcementLearningEnvironmentDiscrete.setterminalstates! — Function.

setterminalstates!(mdp, range)

Sets mdp.isterminal[range] .= 1, empties the table of transition probabilities for terminal states and sets the reward for all actions in the terminal state to the same value.

ReinforcementLearningEnvironmentDiscrete.setwall!

ReinforcementLearningEnvironmentDiscrete.treeMDP — Function.

treeMDP(na, depth; init = "random", branchingfactor = 3)

Returns a tree structured MDP with na actions and depth of the tree. If init is random, the branchingfactor determines how many possible states a (action, state) pair has. If init = "deterministic" the branchingfactor = na.