Docstrings

Paillier.Ciphertext

Paillier.Encoded — Type.

A datatype for a plaintext encoded number. Returned by the encode methods.

Paillier.Encoding — Type.

Encoding(::DataType, ::PublicKey)
Encoding(::DataType, ::PublicKey, base::Int64)

A datatype for describing an encoding scheme.

The public key is included as the encoding is effected by the maximum representable integer which varies with the public_key. Although I could be convinced to change this.

Examples

Setting a base value is optional:

julia> encoding = Encoding(Float64, public_key, 64)

Full example showing homomorphic operations on floating point numbers:

julia> keysize = 2048
julia> publickey, privatekey = generate_paillier_keypair(keysize)
julia> encoding = Encoding(Float32, publickey)
julia> a = Float32(π)
julia> enc1 = encode_and_encrypt(a, encoding)
julia> decrypt_and_decode(privatekey, enc1)
3.1415927f0
julia> enc1.exponent
-6
julia> b = 100
julia> enc2 = encode_and_encrypt(b, encoding)
julia> decrypt_and_decode(privatekey, enc1 + enc2)
103.141594f0
julia> decrypt_and_decode(privatekey, enc1 - 20.0)
-16.858408f0

Paillier.Encrypted — Type.

Encrypted(ciphertext, public_key)

An Encrypted is the Paillier.jl library's low level encrypted type. This simple object that includes the ciphertext, public_key and tracks whether obfuscation has occurred (assumed as false if not provided).

Paillier.EncryptedArray — Type.

EncryptedArray

A vector version of EncryptedNumber.

Paillier.EncryptedNumber — Type.

EncryptedNumber(::Encrypted, ::Encoding, exponent::Int64)
EncryptedNumber(::Encoded, ::PublicKey)

Datatype for representing an encrypted (and Encoded) number.

Examples

julia> EncryptedNumber(encoded_number, public_key)

Paillier.Paillier — Module.

module Paillier

Examples

julia> pub, priv = generate_paillier_keypair(1024)
julia> c = encrypt(pub, 10)
julia> decrypt(priv, add(pub, c, encrypt(pub, 20)))
30

Paillier.PrivateKey

Paillier.PublicKey

Paillier._encrypt_encoded

Paillier._multiply_via_inversion

Paillier.decode — Function.

decode(encoded::Encoded)
decode(encoded::BigInt, exponent::Int64, encoding::Encoding)

The inverse of encode, computes the decoding of the Encoded integer representation.

Paillier.decrease_exponent_to — Function.

decrease_exponent_to(enc, new_exponent)

Compute an EncryptedNumber with the same value but a lower exponent.

We multiply the encoded value by the Encoding.base and decrement the EncryptedNumber.exponent - such that the decoded value remains the same.

We can keep ratcheting down the EncryptedNumber's exponent until the encoded integer overflows. This overflow may not be caught.

There is also a (much faster) version that acts on EncodedNumbers.

Paillier.decrypt — Function.

decrypt(private_key, ciphertext::Ciphertext)
decrypt(private_key, ciphertext::Encrypted)

The inverse of encrypt, uses the private_key to decrypt an encrypted message (ciphertext) to a positive integer less than public_key.n. The result is always a BigInt.

Examples

julia> publickey, privatekey = generate_paillier_keypair(128)
julia> ciphertext = encrypt(publickey, 10)
julia> decrypt(privatekey, ciphertext)
10

Paillier.decrypt_and_decode — Function.

decrypt_and_decode(privatekey::PrivateKey, enc::EncryptedNumber)

Decrypt an EncryptedNumber using the given PrivateKey and decode it using the EncryptedNumber's encoding.

Paillier.encode — Function.

encode(::Number, ::Encoding)
encode(::Number, ::Encoding, exponent::Int64)

Encode a number but don't encrypt it for the given Encoding - producing an Encoded. See encode_and_encrypt for a method that also encrypts. If the exponent is not provided we attempt to match the precision of the passed julia type. See decode to go the other direction.

Paillier.encode_and_encrypt — Function.

encode_and_encrypt(plaintext::Number, encoding::Encoding)

Encode the plaintext number using given encoding and encrypt using the PublicKey from the encoding.

encode_and_encrypt(xs::Array{<:Number}, encoding::Encoding)
encode_and_encrypt(xs::Array{<:Number}, encoding::Encoding, exponent::Int64)

Create an EncryptedArray of your plaintext numbers.

Paillier.encrypt — Function.

encrypt(public_key, message)

Encrypt a message with a given public key and return an Encrypted. The message must be a positive integer under public_key.n - following Julia's Int64 type larger numbers will wrap around and not throw an error.

See decrypt if you'd like your original message back.

Paillier.encrypt_raw — Function.

encrypt_raw(public_key, message)

Internal version of encrypt that returns the raw Ciphertext - which is just a BigInt. Note this includes obfuscation so a directly encrypted unsigned integer is safe to share.

Paillier.eval

Paillier.find_ea — Function.

A = find_ea(As) returns the first EncryptedArray among the arguments.

Paillier.generate_paillier_keypair

Paillier.include

Paillier.intrep

Paillier.match_exponents

Paillier.max_int — Function.

max_int(::PublicKey)

The maximum signed integer for our encoding system. We use a full third of the range for overflow detection.

Paillier.n_bit_random_number

Paillier.nbit_prime_of_size

Paillier.obfuscate — Function.

obfuscate(encrypted)

Salt the Encrypted with a new random number. Required before sharing ciphertexts with another party.

Paillier.random_lt_n

Paillier.raw_add

Paillier.raw_multiply