Docstrings

CompScienceMeshes.AbstractMapper

CompScienceMeshes.CompScienceMeshes

CompScienceMeshes.FlippedMesh

CompScienceMeshes.FlippedMeshCellIterator

CompScienceMeshes.GSubdMesh — Function.

GSubdMesh(mesh::Mesh)

Given a linear polygon mesh, construct a data structure for subdivision surfaces

CompScienceMeshes.JacobMatrix33

CompScienceMeshes.Jacob_inv

CompScienceMeshes.Loop_subdivision — Function.

Loop_subdivsion(mesh) -> refinement

Construct a refinement of mesh by Loop subdivision scheme. Every face is subdived into four small faces and use weights to smooth the surface.

Only defined for 2D meshes.

CompScienceMeshes.Mesh

CompScienceMeshes.MeshPointNM

CompScienceMeshes.Pt

CompScienceMeshes.RectangleChart — Type.

N: universe dimension D: manifold dimension T: coordinate type

CompScienceMeshes.RectangleNBD

CompScienceMeshes.ReferenceSimplex — Type.

ReferenceSimplex{Dimension, CoordType, NumVertices}

This domain is defined to bootstrap the quadrature generation strategy. The generic definition of numquads on a chart pulls back to the domain. For a limit set of reference domains, explicit quadrature rules are defined. The weights and points are then pushed forward to the configuaration space element over which integration is desired.

For more details see the implementation in quadpoints.jl

CompScienceMeshes.SEdge

CompScienceMeshes.SElement

CompScienceMeshes.SVertex

CompScienceMeshes.SW

CompScienceMeshes.SW1

CompScienceMeshes.SW10

CompScienceMeshes.SW11

CompScienceMeshes.SW12

CompScienceMeshes.SW13

CompScienceMeshes.SW14

CompScienceMeshes.SW15

CompScienceMeshes.SW16

CompScienceMeshes.SW17

CompScienceMeshes.SW18

CompScienceMeshes.SW19

CompScienceMeshes.SW2

CompScienceMeshes.SW20

CompScienceMeshes.SW21

CompScienceMeshes.SW22

CompScienceMeshes.SW23

CompScienceMeshes.SW24

CompScienceMeshes.SW25

CompScienceMeshes.SW26

CompScienceMeshes.SW27

CompScienceMeshes.SW28

CompScienceMeshes.SW29

CompScienceMeshes.SW3

CompScienceMeshes.SW30

CompScienceMeshes.SW4

CompScienceMeshes.SW5

CompScienceMeshes.SW6

CompScienceMeshes.SW7

CompScienceMeshes.SW8

CompScienceMeshes.SW9

CompScienceMeshes.SX

CompScienceMeshes.SX1

CompScienceMeshes.SX10

CompScienceMeshes.SX11

CompScienceMeshes.SX12

CompScienceMeshes.SX13

CompScienceMeshes.SX14

CompScienceMeshes.SX15

CompScienceMeshes.SX16

CompScienceMeshes.SX17

CompScienceMeshes.SX18

CompScienceMeshes.SX19

CompScienceMeshes.SX2

CompScienceMeshes.SX20

CompScienceMeshes.SX21

CompScienceMeshes.SX22

CompScienceMeshes.SX23

CompScienceMeshes.SX24

CompScienceMeshes.SX25

CompScienceMeshes.SX26

CompScienceMeshes.SX27

CompScienceMeshes.SX28

CompScienceMeshes.SX29

CompScienceMeshes.SX3

CompScienceMeshes.SX30

CompScienceMeshes.SX4

CompScienceMeshes.SX5

CompScienceMeshes.SX6

CompScienceMeshes.SX7

CompScienceMeshes.SX8

CompScienceMeshes.SX9

CompScienceMeshes.SY

CompScienceMeshes.SY1

CompScienceMeshes.SY10

CompScienceMeshes.SY11

CompScienceMeshes.SY12

CompScienceMeshes.SY13

CompScienceMeshes.SY14

CompScienceMeshes.SY15

CompScienceMeshes.SY16

CompScienceMeshes.SY17

CompScienceMeshes.SY18

CompScienceMeshes.SY19

CompScienceMeshes.SY2

CompScienceMeshes.SY20

CompScienceMeshes.SY21

CompScienceMeshes.SY22

CompScienceMeshes.SY23

CompScienceMeshes.SY24

CompScienceMeshes.SY25

CompScienceMeshes.SY26

CompScienceMeshes.SY27

CompScienceMeshes.SY28

CompScienceMeshes.SY29

CompScienceMeshes.SY3

CompScienceMeshes.SY30

CompScienceMeshes.SY4

CompScienceMeshes.SY5

CompScienceMeshes.SY6

CompScienceMeshes.SY7

CompScienceMeshes.SY8

CompScienceMeshes.SY9

CompScienceMeshes.SegmentedAxis — Type.

This type conforms to the mesh interface but is specialised for the case of a segment of the real axis subdived in equally sized intervals. Typical use is a discretisation of the time axis.

CompScienceMeshes.Simplex

CompScienceMeshes.SphereChart

CompScienceMeshes.SphereNeighborhood

CompScienceMeshes.ThetaPhiPlane

CompScienceMeshes.Trianglegauss105a

CompScienceMeshes.Trianglegauss105b

CompScienceMeshes.Trianglegauss105w

CompScienceMeshes.Trianglegauss120a

CompScienceMeshes.Trianglegauss120b

CompScienceMeshes.Trianglegauss120w

CompScienceMeshes.Trianglegauss12a

CompScienceMeshes.Trianglegauss12b

CompScienceMeshes.Trianglegauss12c

CompScienceMeshes.Trianglegauss12w

CompScienceMeshes.Trianglegauss13a

CompScienceMeshes.Trianglegauss13b

CompScienceMeshes.Trianglegauss13c

CompScienceMeshes.Trianglegauss13w

CompScienceMeshes.Trianglegauss1a

CompScienceMeshes.Trianglegauss1b

CompScienceMeshes.Trianglegauss1c

CompScienceMeshes.Trianglegauss1w

CompScienceMeshes.Trianglegauss36a

CompScienceMeshes.Trianglegauss36b

CompScienceMeshes.Trianglegauss36w

CompScienceMeshes.Trianglegauss3a

CompScienceMeshes.Trianglegauss3b

CompScienceMeshes.Trianglegauss3c

CompScienceMeshes.Trianglegauss3w

CompScienceMeshes.Trianglegauss400a

CompScienceMeshes.Trianglegauss400b

CompScienceMeshes.Trianglegauss400w

CompScienceMeshes.Trianglegauss4a

CompScienceMeshes.Trianglegauss4b

CompScienceMeshes.Trianglegauss4c

CompScienceMeshes.Trianglegauss4w

CompScienceMeshes.Trianglegauss6a

CompScienceMeshes.Trianglegauss6b

CompScienceMeshes.Trianglegauss6c

CompScienceMeshes.Trianglegauss6w

CompScienceMeshes.Trianglegauss78a

CompScienceMeshes.Trianglegauss78b

CompScienceMeshes.Trianglegauss78w

CompScienceMeshes.Trianglegauss7a

CompScienceMeshes.Trianglegauss7b

CompScienceMeshes.Trianglegauss7c

CompScienceMeshes.Trianglegauss7w

CompScienceMeshes.Trianglegauss900a

CompScienceMeshes.Trianglegauss900b

CompScienceMeshes.Trianglegauss900w

CompScienceMeshes.WeightPointValue

CompScienceMeshes.__init__

CompScienceMeshes._normals

CompScienceMeshes.barycentric — Function.

Return the barycentric coordinates of mp

CompScienceMeshes.barycentric_refinement — Function.

barycentric refinement(mesh) -> refined_mesh

Create the mesh obtained by inserting an extra vertex in the barycenters of all cells and recusively creating fine cells by connecting the barycenter of a k-cell to the already constructed refined (k-1)-cells on its boundary.

CompScienceMeshes.barytocart — Function.

barytocart(simplex, uv)

Returns the point in the simplex with barycentric coordinates uv

CompScienceMeshes.bisecting_refinement — Function.

bisecting_refinement(mesh) -> refinement

Construct a refinement of mesh by inserting a new vertex on every existing edge. Every face is subdived in four small faces.

Only defined for 2D meshes.

CompScienceMeshes.boundary — Function.

boundary(mesh)

Returns the boundary of mesh as a mesh of lower dimension.

CompScienceMeshes.boundingbox — Function.

Returns the boundingbox of a patch in terms of its center and halfsize.

function boundingbox{U,D,C,N,T}(p::Simplex{U,D,C,N,T}) -> center, halfsize

CompScienceMeshes.boxspline_func

CompScienceMeshes.boxspline_func_der

CompScienceMeshes.boxspline_func_der2

CompScienceMeshes.cartesian

CompScienceMeshes.carttobary — Function.

carttobary(simplex, point) -> barycoords

Compute the barycentric coordinates on 'simplex' of 'point'.

CompScienceMeshes.cellarray

CompScienceMeshes.cellpairs — Function.

pairs = cellpairs(mesh, edges, dropjunctionpair=false)

Given a mesh and set of oriented edges from that mesh (as generated by skeleton), cellpairs will generate a 2 x K matrix, where K is the number of pairs and each column contains a pair of indices in the cell array of mesh that have one of the supplied edges in common.

Returns an array of pairs of indices, each pair corresponding to a pair of adjacent faces.

(If the mesh is oriented, the first row of facepairs will contain indices to the cell for which the corresponding edge has a positive relative orientation.

If a edge lies on the boundary of the mesh, and only has one neighboring cell, the second row of facepairs will contain -k with k the local index of the corresponding edge in its neighboring triangle.

If an edge has more than two neighboring cells (i.e. the edge is on a junction), all possible pairs of cells that have the junction edge in common are supplied. if dropjunctionpair == false then one of the possible pairs of cells is not recorded. This is done to avoid the creation of linearly dependent basis functions in the construction of boundary element methods for Maxwell's equations.)

CompScienceMeshes.cells — Function.

cells(mesh)

Return an iterable collection containing the cells making up the mesh.

CompScienceMeshes.celltype — Function.

celltype(mesh)

Returns the type of the index tuples stored in the mesh.

CompScienceMeshes.center

CompScienceMeshes.chart — Function.

chart(mesh, cell) -> cell_chart

Return a chart describing the supplied cell of mesh.

CompScienceMeshes.clipconvex!

CompScienceMeshes.connectivity — Function.

connectivity(faces, cells, op=sign)

Create a sparse matrix D of size numcells(cells) by numcells(faces) that contiains the connectivity info of the mesh. In particular D[m,k] is op(r) where r is the local index of face k in cell m. The sign of r is positive or negative depending on the relative orientation of face k in cell m.

For op=sign, the matrix returned is the classic connectivity matrix, i.e. the graph version of the exterior derivative.

CompScienceMeshes.coordtype — Function.

coordtype(mesh)

Returns eltype(vertextype(mesh))

coordtype(simplex)

Return coordinate type used by simplex.

CompScienceMeshes.dimension — Function.

dim = dimension(mesh)

Returns the dimension of the mesh. Note that this is the dimension of the cells, not of the surrounding space.

dimension(simplex)

Return the manifold dimension of the simplex.

dimension(simplex)

Return the manifold dimension of the simplex.

CompScienceMeshes.domain

CompScienceMeshes.euclidianbasis — Function.

euclidian_basis(type, dim)

Returns the origin and default unit vectors for Euclidian space of dimension dim

CompScienceMeshes.eval

CompScienceMeshes.find_edges

CompScienceMeshes.find_neighbor — Function.

find_neighbor(faces,edges,F,EdgesIndices,orientation,Sedges)

Given a face find out the neighbor elements (share edge) and vertices.

CompScienceMeshes.find_neighbor2

CompScienceMeshes.flip — Function.

flip(cell)

Change the orientation of a cell by interchanging the first to indices.

CompScienceMeshes.flipmesh

CompScienceMeshes.flipmesh! — Function.

flipmesh!(mesh)

Change the orientation of a mesh

CompScienceMeshes.fliporientation — Function.

fliporientation(mesh)

Returns a mesh of opposite orientation.

CompScienceMeshes.fliporientation! — Function.

fliporientation(mesh)

Changes the mesh orientation inplace. If non-orientatble, undefined.

CompScienceMeshes.gauss_points

CompScienceMeshes.gauss_quad_linear

CompScienceMeshes.get_shape_curl

CompScienceMeshes.getcommonedge

CompScienceMeshes.getelementVertices

CompScienceMeshes.inclosure_gpredicate — Function.

Geometric predicate for determining in log(N) complexity if a the image of a chart is in the closure of mesh γ.

CompScienceMeshes.include

CompScienceMeshes.index — Function.

index(i1, i2, ...) -> ids

Create a tuple of vertex indices.

CompScienceMeshes.interior_tpredicate — Function.

Creates a predicate that can be used to check wheter an edge is interior to a surface (true) or on its boundary (false). This predicate is based on combinatorics. In particular it expects as argument a tuple of indices pointing into the vertex buffer of mesh.

CompScienceMeshes.intersection — Function.

intersect(chartA, chartB) -> [chart1, chart2, ...]

Compute the intersection of two charts of equal dimension.

Compute an array of charts such that the disjoint union of those charts produces the intersection of the two charts provided as inputs. In particular the sum of integrals over the returned charts will equal the integral over the intersection of the two given charts.

intersection(triangleA, triangle B)

ATTENTION: currently the implementation for triangles assumes that one of the triangles is contained in the other.

CompScienceMeshes.intersectlines — Function.

intersectline(a,b,p,q)

Computes the intersection of the lines (in a 2D space) defined by points [a,b] and [p,q]

CompScienceMeshes.irreg_shapefun

CompScienceMeshes.isinclosure — Function.

isinclosure(simplex, point) -> Bool

Determine whether point is in the closure of simplex. False positives are possible for points just outside of the simplex.

CompScienceMeshes.isinside — Function.

isinside(chart, point) -> Bool

Returns true is the given point is in the image of the given chart, false otherwise.

CompScienceMeshes.jacobian — Function.

A number defines a neighborhood in euclidian space

CompScienceMeshes.leftof — Function.

inside(p,a,b)

Determines is p is on the interior side of the segment b of the boundary, assuming that the boundary is oriented counter-clockwise.

CompScienceMeshes.legendre

CompScienceMeshes.load_gid_mesh — Function.

load_gid_mesh(filename) -> mesh

load_gid_mesh(stream) ->mesh

CompScienceMeshes.map_ids

CompScienceMeshes.mapper

CompScienceMeshes.matrix_A

CompScienceMeshes.mesh — Function.

mesh(type, mdim, udim=mdim+1)

Returns an empty mesh with coordtype equal to type, of dimension mdim and embedded in a universe of dimension udim

CompScienceMeshes.meshcircle

CompScienceMeshes.meshcuboid — Function.

meshcuboid(width, height, length, delta)

Creates a mesh for a cuboid of width (along the x-axis) width, height (along the y-axis) height and length (along the z-axis) length by parsing a .geo script incorporating these parameters into the GMSH mesher.

The target edge size is delta.

CompScienceMeshes.meshrectangle — Function.

meshrectangle(width, height, delta, udim)

Create a mesh for a rectangle of width (along the x-axis) width and height (along the y-axis) height.

The target edge size is delta and the dimension of the embedding universe is udim (>= 2).

The mesh is oriented such that the normal is pointing down. This is subject to change.

CompScienceMeshes.meshsegment

CompScienceMeshes.meshsphere — Function.

meshsphere(radius, delta)

Create a mesh of a sphere of radius radius by parsing a .geo script incorporating these parameters into the GMSH mesher.

The target edge size is delta.

CompScienceMeshes.minmaxdist

CompScienceMeshes.mirror — Function.

mirror(vertex, normal, anchor)

Mirror vertex across a plane defined by its normal and a containing point.

CompScienceMeshes.mirrormesh

CompScienceMeshes.mirrormesh!

CompScienceMeshes.neighborhood

CompScienceMeshes.normal

CompScienceMeshes.numcells — Function.

numcells(mesh)

Returns the number of cells in the mesh.

CompScienceMeshes.numvertices — Function.

numvertices(mesh)

Returns the number of vertices in the mesh.

Note: this is the number of vertices in the vertex buffer and might include floatin vertices or vertices not appearing in any cell. In other words the following is not necessarily true:

    numvertices(mesh) == numcells(skeleton(mesh,0))

CompScienceMeshes.octree — Function.

Store the k-cells of a mesh in an octree.

function octree{U,D,T}(mesh::Mesh{U,D,T}, kcells::Array{Int,2})

CompScienceMeshes.overlap — Function.

Compute whether two flat patches of the same dimension overlap or not

Compute whether two segments in 3D space overlap

Compute whether two triangles in 3D space overlap

CompScienceMeshes.overlap_gpredicate — Function.

overlap_gpredicate(γ::Mesh) -> (patch -> Bool)

Create a predicate that for a given patch determinees if it overlaps with the provided target mesh γ.

CompScienceMeshes.parametric

CompScienceMeshes.patch

CompScienceMeshes.pickmatrx_reg

CompScienceMeshes.pickmtrx_irreg

CompScienceMeshes.point — Function.

point(xs...)

Create point of default type (double precision coordinates)

point(type, xs...)

Create point of default type and supplied precision for the coordinates

CompScienceMeshes.pow

CompScienceMeshes.quadpoints — Function.

pw = quadpoints(chart, rule)

Returns a collection of (point, weight) tuples corresponding to the numerical quadrature rule defined on the domain of chart. The weights returned take into account the Jacobian determinant resulting from mapping from the reference domain to the configuration space.

Functions can be integrated like:

PW = quadpoints(chart, rule)
I = sum(pw[2]*f(pw[1]) for pw in PW)

quadpoints(refspace, charts, rules)

Computed a matrix of vectors containing (weight, point, value) triples that can be used in numerical integration over the elements described by the charts. Internally, this method used quadpoints(chart, rule) to retrieve the points and weights for a certain quadrature rule over chart.

CompScienceMeshes.read_TRI_mesh — Function.

read_TRI_mesh(filename) -> mesh::Mesh

read_TRI_mesh(mesh_filename) -> mesh::Mesh

Imports a surface mesh (stored in an ASCII file named mesh_filename) into a Mesh object (i.e. node list and element list).

NOTE: The contents of mesh_filename must include the file extension, and the file must be stored in the current directory.

CompScienceMeshes.read_gmsh_mesh — Function.

read_gmsh_mesh(filename) -> mesh::Mesh

read_gmsh_mesh(iostream) -> mesh::Mesh

Reads the mesh nodes and elements stored in the input .msh file (io, output by GMSH) into arrays of node vectors and vertex vectors respectively.

Returns an object mesh of type Mesh, comprising both vector arrays.

CompScienceMeshes.readmesh — Function.

readmesh(filename)

Reads a mesh in in format from filename. The format follows:

1
V C
x1_1    x1_2    ... x1_U
x2_1    x2_2    ... x2_U
...
xV_1    xV_2    ... xV_U
i1_1    i1_2    ... i1_D1
i2_1    i2_2    ... i2_D1
...
iC_1    iC_2    ... iC_D1

where U is the universedimension of the mesh, D1 the dimension of the mesh plus one, V the number of vertices, and C the number of cells in the mesh.

CompScienceMeshes.reg_shapefun

CompScienceMeshes.relorientation

CompScienceMeshes.restriction — Function.

restriction(submesh, supermesh)

Computes the restriction matrix relative to a submesh submesh of supermesh.

The restriction matrix has size (m,n), where

m == numcells(submesh)
n == numcells(supermesh)

It has entries 1 at location [i,j] iff cell i of submesh equals cell j of supermesh. The remaining entries are zero.

This matrix is referred to as the restriction matrix because if it acts on an array of samples taken at the cells of supermesh is selects out the samples in the cells that are retained in submesh, taking into account any renumbering. Likewise, its transpose is sometimes referred to as the extension-by-zero operator because it maps arrays of samples taken in the cells of submesh into an array of samples taken in the cells of supermesh by inserting zeros at cells that were not retained in submesh.

CompScienceMeshes.ring

CompScienceMeshes.rings

CompScienceMeshes.rotate

CompScienceMeshes.rotate!

CompScienceMeshes.shape_function

CompScienceMeshes.shape_function_der

CompScienceMeshes.shape_function_der2

CompScienceMeshes.shapeder

CompScienceMeshes.shapeders

CompScienceMeshes.shapefuns

CompScienceMeshes.simplex — Function.

simplex(vertices)
simplex(v1, v2, ...)
simplex(vertices, Val{D})

Build a D-dimensional simplex. The vertices can be passed in an array (static or dynamic), or supplied separately. If the length of the array is not part of its type, the speed of the construction can be improved by supplying an extra Val{D} argument. In case it is not clear from the context whether the vertex array is dynamically or statically sized, use the third form as it will not incur notable performance hits.

Note that D is the dimension of the simplex, i.e. the number of vertices supplied minus one.

CompScienceMeshes.skeleton — Function.

skeleton(mesh, dim)

Returns a mesh comprising the dim-dimensional sub cells of mesh. For example to retrieve the edges of a given surface mesh,

edges = skelton(mesh, 1)

skeleton(pred, mesh, dim)

Like skeleton(mesh, dim), but only cells for which pred(cell) returns true are withheld.

CompScienceMeshes.sqgauss

CompScienceMeshes.subdMesh

CompScienceMeshes.subd_chart

CompScienceMeshes.subd_point

CompScienceMeshes.submesh — Function.

Returns a mesh on the same vertexbuffer as the input mesh. The submesh will be a mesh of dimension k containing all the k-cells that are in mesh and that fulfill the predicate pred.

pred is a function with signature pred(cell) -> Bool returning true if the simplex is to be added to the submesh under construction.

submesh(selection, mesh)

Create a submesh from mesh comprising those elements that overlap with elements from selection. It is assumed that selection and mesh have the same dimension.

CompScienceMeshes.sutherlandhodgman — Function.

sutherlandhodgman(subject, clipper)

Compute the intersection of two coplanar triangles, potentially embedded in a higher dimensional space.

CompScienceMeshes.sutherlandhodgman2d — Function.

sutherlandhodgman2d(subject,clipper)

Computes the intersection of the coplanar triangles subject and clipper.

CompScienceMeshes.tangents

CompScienceMeshes.translate — Function.

translate(mesh, v)

Creates a new mesh by translating mesh over vector v

CompScienceMeshes.translate! — Function.

translate!(mesh, v)

Translates mesh over vector v inplace.

CompScienceMeshes.trgauss — Function.

trgauss(n) -> (u,w)

Returns the n-th triangle quadrature rule. Returns a Matrix u of size (Q,2) with Q the number of quadrature points and a Vector w of size (Q,) containing the quadrature weights.

CompScienceMeshes.triangleGaussA

CompScienceMeshes.triangleGaussB

CompScienceMeshes.triangleGaussW

CompScienceMeshes.universedimension — Function.

universedimension(mesh)

Returns the dimension of the surrounding space. Equals the number of coordinates required to describe a vertex.

universedimension(p)

Return the dimension of the universe in which p is embedded.

CompScienceMeshes.vertexarray

CompScienceMeshes.vertextocellmap — Function.

vertextocellmap(mesh) -> vertextocells, numneighbors

Computed an V×M array vertextocells where V is the number of vertices and M is the maximum number of cells adjacent to any given vertex such that vertextocells[v,i] is the index in the cells of mesh of the ith cell adjacent to teh v-th vertex. numneighbors[v] contains the number of cells adjacent to the v-th vertex.

This method allows e.g. for the efficient computation of the connectivity matrix of the mesh.

CompScienceMeshes.vertextype — Function.

vt = vertextype(mesh)

Returns type of the vertices used to define the cells of the mesh.

CompScienceMeshes.vertices — Function.

vertices(mesh)

Returns an indexable iterable to the vertices of the mesh

CompScienceMeshes.volume — Function.

volume(simplex)

Return the volume of the simplex.

A tuple of points, aka an interval behaves trivially like a chart

CompScienceMeshes.weld — Function.

weld(mesh1, mesh2, ...) -> welded_mesh

Build a mesh by welding or pasting together the inputs. Vertices from different meshes that coincide up to the tolerance will be merged into one. The order cells appear in the output mesh is equal to the order in the inputs.

CompScienceMeshes.writemesh — Function.

writemesh(mesh, filename)

Write mesh to filename in the in format (see readmesh).