Docstrings

SortedVectors.AssumeSorted — Type.

Flag for indicating that

the argument should be assumed to be sorted, and this should not be checked,

the argument vector will not be modified later.

Not exported.

SortedVectors.CheckSorted — Type.

Flag for indicating that

sorting should be verified,

the argument vector will not be modified later.

Not exported.

SortedVectors.SortedVector — Type.

SortedVector([lt=isless], xs)

Sort xs by lt (which defaults to isless) and wrap in a SortedVector. For reverse sorting, use !lt.

SortedVector(SortedVectors.CheckSorted(), lt, sorted_contents)

Checks that the vector is sorted, throws an ArgumentError if it isn't. This is a relatively cheap operation if the vector is supposed to be sorted but this should be checked. Caller should copy the sorted_contents if they are mutable and may be modified.

SortedVector(SortedVectors.AssumeSorted(), lt, sorted_contents)

Unchecked, unsafe constructor. Use only if you are certain that sorted_contents is sorted according to lt, otherwise results are undefined. copy the sorted_contents if they are mutable and may be modified.

SortedVectors.SortedVectors

SortedVectors.cut — Function.

cut(x, breaks; left, right)

Return i such that breaks[i] < x ≤ breaks[i + 1], where < is the sorting from the second argument.

When x ≤ breaks[1], left (0 by default) is used, and right (default: length) for x > breaks[end].

SortedVectors.eval

SortedVectors.include