Docstrings

MonteCarloObservable.@diskobs — Macro.

Convenience macro for generating a "disk observable" (inmemory=false) from a vector of measurements.

MonteCarloObservable.@obs — Macro.

Convenience macro for generating an Observable from a vector of measurements.

MonteCarloObservable.DiskObservable

MonteCarloObservable.Jackknife — Module.

Jackknife errors for (non-linear) functions of uncertain data, i.e. g(<a>,<b>,...)

Based on https://github.com/ararslan/Jackknife.jl.

MonteCarloObservable.MonteCarloObservable — Module.

A package for handling observables in a Markov Chain Monte Carlo simulation.

See http://github.com/crstnbr/MonteCarloObservable.jl for more information.

MonteCarloObservable.Observable — Type.

A Markov Chain Monte Carlo observable.

MonteCarloObservable.R_function — Function.

Groups datapoints in bins of varying size bs. Returns the used binsizes bss, the error coefficient function R(bss) (Eq. 3.20), and the averaged error coefficients <R>(bss). The function should feature a plateau, i.e. R(bs_p) ~ R(bs) for bs >= bs_p. (Fig. 3.3)

Optional keyword min_nbins. Only bin sizes used that lead to at least min_nbins bins.

MonteCarloObservable.R_value — Function.

Groups datapoints in bins of fixed binsize and returns error coefficient R. (Eq. 3.20)

MonteCarloObservable.Rplateaufinder — Function.

Estimates the error coefficient R by considering <R>(bss) (averaged R values).

Returns estimated start of plateau (bin size), estimate for R, and convergence (boolean).

MonteCarloObservable.SUPPORTED_TYPES

MonteCarloObservable.TimeSeriesSerializer — Type.

Thin wrapper type used in JLD.writeas to dump the contained vector as a higher-dimensional matrix.

MonteCarloObservable._add!

MonteCarloObservable._frommemory

MonteCarloObservable._getindex_ts_chunk

MonteCarloObservable._init! — Function.

_init!(obs)

Initialize non-external fields of observable obs.

MonteCarloObservable._println_body

MonteCarloObservable._println_header

MonteCarloObservable.add! — Function.

Add measurements to an observable.

add!(obs::Observable{T}, measurement::T; verbose=false)
add!(obs::Observable{T}, measurements::AbstractVector{T}; verbose=false)

MonteCarloObservable.binning_error — Function.

binning_error(X)

Estimate of the one-sigma error of the time series's mean. Respects correlations between measurements through binning analysis.

Note that this is not the same as Base.std(X), not even for uncorrelated measurements.

For more details, see MonteCarloObservable.Rplateaufinder.

binning_error(X, binsize)

Estimate of the one-sigma error of the time series's mean. Respect correlations between measurements through binning analysis, using the given binsize (i.e. assuming independence of bins, Eq. 3.18 basically).

MonteCarloObservable.binning_error_from_R

MonteCarloObservable.binning_error_naive — Function.

binning_error_naive(X, min_nbins=50)

Estimate of the one-sigma error of the observable's mean. Respects correlations between measurements through binning analysis.

Strategy: just take largest R value considering an upper limit for bin size (min_nbins)

MonteCarloObservable.binning_error_with_convergence — Function.

Returns one sigma error and convergence flag (boolean).

Base.eltype — Function.

eltype(obs::Observable{T})

Returns the type T of a measurment of the observable.

Base.error — Function.

Estimate of the one-sigma error of the observable's mean. Respects correlations between measurements through binning analysis.

Note that this is not the same as Base.std(timeseries(obs)), not even for uncorrelated measurements.

See also mean(obs).

error(g::Function, x::AbstractMatrix)

Compute the jackknife estimate of the one sigma error of g(<a>,<b>,...), where g is given as a function that computes a point estimate (scalar) when passed a matrix x. Columns of x are time series of the random variables.

For more details, see also `leaveoneout.

MonteCarloObservable.error_frommemory

MonteCarloObservable.error_naive — Function.

Estimate of the one-sigma error of the observable's mean. Respects correlations between measurements through binning analysis.

Strategy: just take largest R value considering an upper limit for bin size (min_nbins)

MonteCarloObservable.error_with_convergence — Function.

Returns one sigma error and convergence flag (don't trust it!).

MonteCarloObservable.eval

MonteCarloObservable.export_error — Function.

export_error(obs::Observable{T}[, filename::AbstractString, group::AbstractString;])

Export one-sigma error estimate and convergence flag.

MonteCarloObservable.export_result — Function.

export_results(obs::Observable{T}[, filename::AbstractString, group::AbstractString; timeseries::Bool=false])

Export result for given observable nicely to JLD.

Will export name, number of measurements, estimates for mean and one-sigma error. Optionally (timeseries==true) exports the full time series as well.

MonteCarloObservable.fileext — Function.

fileext(filepath::AbstractString)

Extracts lowercase file extension from given filepath. Extension is defined as "everything after the last dot".

Base.flush — Function.

flush(obs::Observable)

This is the crucial function if inmemory(obs) == false. It updates the time series on disk. It is called from add! everytime the alloc limit is reached (overflow).

You can call the function manually to save an intermediate state.

MonteCarloObservable.getfrom

Base.getindex — Function.

getindex(obs::Observable{T}, args...)

Get an element of the time series of the observable.

MonteCarloObservable.getindex_fromfile

MonteCarloObservable.getindexrange_fromfile

MonteCarloObservable.include

MonteCarloObservable.inmemory — Function.

Checks wether the observable is kept in memory (vs. on disk).

MonteCarloObservable.isconverged — Function.

isconverged(X)

Checks if the estimation of the one sigma error is converged.

Returns true once the mean R value is converged up to 0.1% accuracy. This corresponds to convergence of the error itself up to ~3% (sqrt).

Base.isempty — Function.

Returns true if the observable hasn't been measured yet.

MonteCarloObservable.isinmemory — Function.

Checks wether the observable is kept in memory (vs. on disk).

MonteCarloObservable.iswithinerrorbars — Function.

iswithinerrorbars(a, b, δ[, print=false])

Checks whether numbers a and b are equal up to given error δ. Will print x ≈ y + k·δ for print=true.

Is equivalent to isapprox(a,b,atol=δ,rtol=zero(b)).

iswithinerrorbars(A::AbstractArray{T<:Number}, B::AbstractArray{T<:Number}, Δ::AbstractArray{<:Real}[, print=false])

Elementwise check whether A and B are equal up to given real error matrix Δ. Will print A ≈ B + K.*Δ for print=true.

Base.iterate — Function.

iterate(iter [, state]) -> Tuple{Array{Complex{Float64},1},Int64}

Implementation of Julia's iterator interface

MonteCarloObservable.jackknife_error — Function.

jackknife_error(g::Function, obs1, ob2, ...)

Computes the jackknife one sigma error of g(<obs1>, <obs2>, ...) by performing a "leave-one-out" analysis.

The function g(x) must take one matrix argument x, whose columns correspond to the time series of the observables, and produce a scalar (point estimate).

Example:

g(x) = @views mean(x[:,1])^2 - mean(x[:,2].^2) followed by jackknife_error(g, obs1, obs2). Here x[:,1] is basically timeseries(obs1) and x[:,2] is timeseries(obs2).

MonteCarloObservable.jltype — Function.

Basically a limited but much more secure version of eval(Meta.parse(s)).

Base.length — Function.

Length of observable's time series.

MonteCarloObservable.listobs — Function.

List all observables in a given file and HDF5 group.

MonteCarloObservable.loadobs — Function.

loadobs(filename::AbstractString, entryname::AbstractString)

Load complete representation of an observable from JLD file.

See also saveobs.

MonteCarloObservable.loadobs_frommemory — Function.

loadobs_frommemory(filename::AbstractString, group::AbstractString)

Create an observable based on a memory dump (inmemory==false).

Statistics.mean — Function.

Mean of the observable's time series.

MonteCarloObservable.mean_frommemory

MonteCarloObservable.name — Function.

Name of the Observable.

Base.ndims — Function.

Number of dimensions of the observable (of one measurement).

Equivalent to ndims(T).

Base.push! — Function.

Add measurements to an observable.

push!(obs::Observable{T}, measurement::T; verbose=false)
push!(obs::Observable{T}, measurements::AbstractArray{T}; verbose=false)

Note that because of internal preallocation this isn't really a push.

MonteCarloObservable.rename — Function.

rename(obs::Observable, name)

Renames the observable.

MonteCarloObservable.reset! — Function.

reset!(obs::Observable{T})

Resets all measurement information in obs.

MonteCarloObservable.rmobs — Function.

Remove an observable.

MonteCarloObservable.saveobs — Function.

saveobs(obs::Observable{T}[, filename::AbstractString, entryname::AbstractString])

Saves complete representation of the observable to JLD file.

Default filename is "Observables.jld" and default entryname is name(obs).

See also loadobs.

Base.size — Function.

Size of the observable (of one measurement).

Statistics.std — Function.

Standard deviation of the observable's time series (assuming uncorrelated data).

See also mean(obs), var(obs), and error(obs).

MonteCarloObservable.tau — Function.

Integrated autocorrelation time (obtained by binning analysis).

See also error(obs).

MonteCarloObservable.timeseries — Function.

Returns the time series of the observable.

If isinmemory(obs) == false it will read the time series from disk.

See also getindex and view.

MonteCarloObservable.timeseries_flat

MonteCarloObservable.timeseries_frommemory — Function.

timeseries_frommemory(filename::AbstractString, group::AbstractString)

Load time series from memory dump (inmemory==false) in HDF5/JLD file.

Will load and concatenate time series chunks. Output will be a vector of measurements.

MonteCarloObservable.timeseries_frommemory_flat — Function.

timeseries_frommemory_flat(filename::AbstractString, group::AbstractString)

Load time series from memory dump (inmemory==false) in HDF5/JLD file.

Will load and concatenate time series chunks. Output will be higher-dimensional array whose last dimension corresponds to Monte Carlo time.

MonteCarloObservable.ts

MonteCarloObservable.ts_flat

MonteCarloObservable.types

Statistics.var — Function.

Variance of the observable's time series (assuming uncorrelated data).

See also mean(obs), std(obs), and error(obs).

var(g::Function, x::AbstractMatrix)

Compute the jackknife estimate of the variance of g(<a>,<b>,...), where g is given as a function that computes a point estimate (scalar) when passed a matrix x. Columns of x are time series of the random variables.

For more details, see also `leaveoneout.

Base.view — Function.

view(obs::Observable{T}, args...)

Get, if possible, a view into the time series of the observable.