Docstrings

AbstractTrees.AbstractShadowTree

AbstractTrees.AbstractTrees

AbstractTrees.AnnotationNode

AbstractTrees.ImplicitChildStates

AbstractTrees.ImplicitIndexStack

AbstractTrees.ImplicitNodeStack — Type.

Keeps a stack of nodes and their corresponding indices. Note that the last node is not explicitly stored in the nodestack, such that length(nodestack) == length(idx_stack)-1 (unless we're at the root in which case both are empty)

AbstractTrees.ImplicitParents

AbstractTrees.ImplicitRootState

AbstractTrees.ImplicitSiblings

AbstractTrees.ImplicitStack

AbstractTrees.IndexedTree

AbstractTrees.Leaves — Type.

Iterator to visit the leaves of a tree, e.g. for the tree

Any[1,Any[2,3]] ├─ 1 └─ Any[2,3] ├─ 2 └─ 3

we will get [1,2,3]

AbstractTrees.ParentLinks

AbstractTrees.PostOrderDFS — Type.

Iterator to visit the nodes of a tree, guaranteeing that children will be visited before their parents.

e.g. for the tree

Any[1,Any[2,3]] ├─ 1 └─ Any[2,3] ├─ 2 └─ 3

we will get [1,2,3,Any[2,3],Any[1,Any[2,3]]]

AbstractTrees.PreOrderDFS — Type.

Iterator to visit the nodes of a tree, guaranteeing that parents will be visited before their children.

Optionally takes a filter function that determines whether the iterator should continue iterating over a node's children (if it has any) or should consider that node a leaf.

e.g. for the tree

Any[Any[1,2],Any[3,4]] ├─ Any[1,2] | ├─ 1 | └─ 2 └─ Any[3,4] ├─ 3 └─ 4

we will get [Any[Any[1,2],Any[3,4]],Any[1,2],1,2,Any[3,4],3,4]

Invalidation

Modifying the underlying tree while iterating over it, is allowed, however, if parents and sibling links are not explicitly stored, the identify of any parent of the last obtained node does not change (i.e. mutation is allowed, replacing nodes is not).

AbstractTrees.RegularTree

AbstractTrees.ShadowTree

AbstractTrees.SiblingLinks

AbstractTrees.StatelessBFS — Type.

Iterator to visit the nodes of a tree, all nodes of a level will be visited before their children

e.g. for the tree

Any[1,Any[2,3]] ├─ 1 └─ Any[2,3] ├─ 2 └─ 3

we will get [Any[1,Any[2,3]],1,Any[2,3],2,3]

WARNING: This is O(n^2), only use this if you know you need it, as opposed to a more standard statefull approach.

AbstractTrees.StoredParents — Type.

Indicates that this tree stores parent links explicitly. The implementation is responsible for defining the parentind function to expose this information.

AbstractTrees.StoredSiblings — Type.

Indicates that this tree stores sibling links explicitly, or can compute them quickly (e.g. because the tree has a (small) fixed branching ratio, so the current index of a node can be determined by quick linear search). The implementation is responsible for defining the relative_state function to expose this information.

AbstractTrees.Subtree

AbstractTrees.Tree

AbstractTrees.TreeCharSet

AbstractTrees.TreeIterator

AbstractTrees.TreeKind

AbstractTrees._print_tree — Function.

Usage

Prints an ASCII formatted representation of the tree to the given io object. By default all children will be printed up to a maximum level of 5, though this valud can be overriden by the maxdepth parameter. The charset to use in printing can be customized using the charset keyword argument.

Examples

julia> print_tree(STDOUT,Dict("a"=>"b","b"=>['c','d']))
Dict{String,Any}("b"=>['c','d'],"a"=>"b")
├─ b
│  ├─ c
│  └─ d
└─ a
   └─ b

julia> print_tree(STDOUT,Dict("a"=>"b","b"=>['c','d']);
        charset = TreeCharSet('+','\\','|',"--"))
Dict{String,Any}("b"=>['c','d'],"a"=>"b")
+-- b
|   +-- c
|   \-- d
\-- a
   \-- b

AbstractTrees.ascend — Function.

Acends the tree, at each node choosing whether or not to continue.
Note that the parent is computed before the callback is exectuted, allowing
modification of the argument to the callback (as long as the overall tree
structure is not altered).

AbstractTrees.childindices

AbstractTrees.children

AbstractTrees.childstates

AbstractTrees.descend — Function.

Descends the tree, at each node choosing the child given by select callback
or the current node if 0 is returned.

AbstractTrees.descend_left

AbstractTrees.eval

AbstractTrees.first_tree

AbstractTrees.firststate

AbstractTrees.get_primary

AbstractTrees.getindexhighest

AbstractTrees.getnode

AbstractTrees.has_children

AbstractTrees.idxtype

AbstractTrees.include

AbstractTrees.isroot

AbstractTrees.joinstate

AbstractTrees.make_annotations

AbstractTrees.make_zip

AbstractTrees.nextind_or_deadend

AbstractTrees.nextsibling

AbstractTrees.nodeequal

AbstractTrees.nodetype

AbstractTrees.parentind

AbstractTrees.parentlinks

AbstractTrees.parentstate

AbstractTrees.prevsibling

AbstractTrees.print_prefix

AbstractTrees.print_tree

AbstractTrees.printnode

AbstractTrees.relative_state

AbstractTrees.rootstate — Function.

Trees must override with method if the state of the root is not the same as the tree itself (e.g. IndexedTrees should always override this method).

AbstractTrees.second_tree

AbstractTrees.siblinglinks

AbstractTrees.stepstate

AbstractTrees.treekind

AbstractTrees.treemap

AbstractTrees.treemap!

AbstractTrees.update_state!

AbstractTrees.zip_min