Docstrings

LasIO.@gen_io — Macro.

Generate IO expressions macro.

LasIO.GeoAsciiParamsTag

LasIO.GeoDoubleParamsTag

LasIO.GeoKeys — Type.

Create GeoKeys from EPSG code. Assumes CRS is projected and in meters.

LasIO.KeyEntry

LasIO.LasHeader

LasIO.LasIO

LasIO.LasPoint — Type.

Abstract type for ASPRS LAS point data record formats 0 - 3

LasIO.LasPoint0 — Type.

ASPRS LAS point data record format 0

source

LasIO.LasPoint1 — Type.

ASPRS LAS point data record format 1

source

LasIO.LasPoint2 — Type.

ASPRS LAS point data record format 2

source

LasIO.LasPoint3 — Type.

ASPRS LAS point data record format 3

source

LasIO.LasPointColor

LasIO.LasPointTime

LasIO.LasVariableLengthRecord — Type.

A LAS "variable length record" - the generic way to store extra user or organization defined binary metadata in LAS files.

LasIO.PointVector — Type.

Custom PointVector struct for memory mapped LasPoints. Inspiration taken from UnalignedVector.jl and extended it with custom indexing and packing.

ColorTypes.RGB — Type.

The RGB color associated with this point

LasIO.SRID

LasIO.__init__

ColorTypes.blue — Function.

The blue image channel value associated with this point

LasIO.classification — Function.

Classification value as defined in the ASPRS classification table.

LasIO.edge_of_flight_line — Function.

If true, it is the last point before the scanner changes direction.

LasIO.epsg_code — Function.

Get the EPSG code of the projection in the header

LasIO.epsg_code! — Function.

Set the projection in the header, without altering the points

LasIO.eval

LasIO.flag_byte — Function.

Flag byte, contains return number, number of returns, scan direction flag and edge of flight line

Flag byte, as represented in the point data, built up from components

LasIO.generate_io

LasIO.generate_read — Function.

Generate read (unpack) method for structs.

LasIO.generate_write — Function.

Generate write (pack) method for structs.

LasIO.gps_time — Function.

Time tag value at which the point was aquired, a Float64; see is_standard_gps for what it represents

Convert DateTime to GPS Float64, as represented in the point data, assumes time is recorded in Adjusted Standard GPS Time; see is_standard_gps

ColorTypes.green — Function.

The green image channel value associated with this point

LasIO.id_geoasciiparamstag

LasIO.id_geodoubleparamstag

LasIO.id_geokeydirectorytag

LasIO.include

LasIO.intensity — Function.

Integer representation of the pulse return magnitude.

LasIO.is_srs — Function.

test whether a vlr is a GeoKeyDirectoryTag, GeoDoubleParamsTag or GeoAsciiParamsTag

LasIO.is_standard_gps — Function.

If true, GPS Time is standard GPS Time (satellite GPS Time) minus 1e9. If false, GPS Time is GPS Week Time.

Note that not all software sets this encoding correctly.

LasIO.is_wkt — Function.

Check if the projection information is in WKT format (true) or GeoTIFF (false)

LasIO.key_point — Function.

If true, this point is considered to be a model key-point.

LasIO.load

LasIO.number_of_returns — Function.

The total number of returns for a given pulse.

LasIO.pointformat

LasIO.pt_src_id — Function.

This value indicates the file from which this point originated.

LasIO.raw_classification — Function.

Raw classification byte, contains classification, synthetic, key point and withheld

Raw classification byte, as represented in the point data, built up from components

LasIO.read_header

LasIO.read_vlr_data

LasIO.readstring

ColorTypes.red — Function.

The red image channel value associated with this point

LasIO.return_number — Function.

The pulse return number for a given output pulse, starting at one.

LasIO.save

LasIO.savebuf

LasIO.scan_angle — Function.

Angle at which the laser point was output, including the roll of the aircraft.

LasIO.scan_direction — Function.

If true, the scanner mirror was traveling from left to right at the time of the output pulse.

LasIO.skiplasf

LasIO.synthetic — Function.

If true, the point was not created from lidar collection

LasIO.update! — Function.

Update the header bounding box and count based on point data

LasIO.user_data — Function.

This field may be used at the user’s discretion.

LasIO.withheld — Function.

If true, this point should not be included in processing

LasIO.write_key_entry

LasIO.writestring

LasIO.xcoord — Function.

X coordinate (Float64), apply scale and offset according to the header

X value (Int32), as represented in the point data, reversing the offset and scale from the header

LasIO.ycoord — Function.

Y coordinate (Float64), apply scale and offset according to the header

Y value (Int32), as represented in the point data, reversing the offset and scale from the header

LasIO.zcoord — Function.

Z coordinate (Float64), apply scale and offset according to the header

Z value (Int32), as represented in the point data, reversing the offset and scale from the header