Docstrings

StructArrays.ArrayInitializer

StructArrays.LazyRow

StructArrays.SkipConstructor

StructArrays.StructArray — Type.

A type that stores an array of structures as a structure of arrays.

Fields:

fieldarrays: a named tuple of arrays. Also fieldarrays(x)

StructArrays.StructArrayInitializer

StructArrays.StructArrays

StructArrays.StructVector

StructArrays.TiedIndices

StructArrays.__init__

StructArrays._dims

StructArrays._reshape

StructArrays._similar

StructArrays._structarray

StructArrays._undef_array

StructArrays.all_types

StructArrays.buildfromschema

StructArrays.collect_empty_structarray

StructArrays.collect_structarray

StructArrays.collect_to_structarray!

StructArrays.createinstance

StructArrays.createtype

StructArrays.default_array

StructArrays.default_initializer

StructArrays.eltypes

StructArrays.eval

StructArrays.fastpermute!

StructArrays.fieldarrays — Function.

fieldarrays(s::StructArray)

Return the field arrays corresponding to the various entry of the struct as a named tuple. If the struct has no names (e.g. a tuple) automatic names are assigned (:x1, :x2, ...).

Examples

julia> s = StructArray(rand(ComplexF64, 4));

julia> fieldarrays(s)
(re = [0.396526, 0.486036, 0.459595, 0.0323561], im = [0.147702, 0.81043, 0.00993469, 0.487091])

StructArrays.fields

StructArrays.finduniquesorted

StructArrays.foreachfield

StructArrays.forward_vec

StructArrays.getfieldindex

StructArrays.grow_to_structarray!

StructArrays.include

StructArrays.iscompatible

StructArrays.iseltype

StructArrays.map_params

StructArrays.map_types

StructArrays.refine_perm!

StructArrays.similar_structarray

StructArrays.sort_int_range_sub_by!

StructArrays.sort_sub_by!

StructArrays.staticschema

StructArrays.tiedindices

StructArrays.to_structarray

StructArrays.tuple_type

StructArrays.uniquesorted

StructArrays.unsafe_createinstance

StructArrays.widenarray

StructArrays.widenstructarray