Docstrings

PolyJuMP.AbstractPoly

PolyJuMP.AbstractPolynomialBasis — Type.

abstract type AbstractPolynomialBasis end

Polynomial basis of a subspace of the polynomials [Section~3.1.5, BPT12].

[BPT12] Blekherman, G.; Parrilo, P. A. & Thomas, R. R. Semidefinite Optimization and Convex Algebraic Geometry. Society for Industrial and Applied Mathematics, 2012.

PolyJuMP.ConstraintDelegate

PolyJuMP.Data

PolyJuMP.FixedPolynomialBasis — Type.

struct FixedPolynomialBasis{PT<:MultivariatePolynomials.AbstractPolynomialLike, PV<:AbstractVector{PT}} <: AbstractPolynomialBasis
    polynomials::PV
end

Polynomial basis with the polynomials of the vector polynomials. For instance, FixedPolynomialBasis([1, x, 2x^2-1, 4x^3-3x]) is the Chebyshev polynomial basis for cubic polynomials in the variable x.

PolyJuMP.MonomialBasis — Type.

struct MonomialBasis{MT<:MultivariatePolynomials.AbstractMonomial, MV<:AbstractVector{MT}} <: AbstractPolynomialBasis
    monomials::MV
end

Monomial basis with the monomials of the vector monomials. For instance, MonomialBasis([1, x, y, x^2, x*y, y^2]) is the monomial basis for the subspace of quadratic polynomials in the variables x, y.

PolyJuMP.NonNegPoly

PolyJuMP.NonNegPolyMatrix

PolyJuMP.Poly — Type.

struct Poly{PB<:AbstractPolynomialBasis} <: AbstractPoly
    polynomial_basis::PB
end

Polynomial variable $v^\top p$ where $v$ is a vector of new decision variables and $p$ is a vector of polynomials for the basis polynomial_basis.

PolyJuMP.PolyConstraint

PolyJuMP.PolyConstraintRef

PolyJuMP.PolyJuMP

PolyJuMP.PolynomialSet

PolyJuMP.ScaledMonomialBasis — Type.

struct ScaledMonomialBasis{MT<:MultivariatePolynomials.AbstractMonomial, MV<:AbstractVector{MT}} <: AbstractPolynomialBasis
    monomials::MV
end

Scaled monomial basis (see [Section 3.1.5, BPT12]) with the monomials of the vector monomials. Given a monomial $x^\alpha = x_1^{\alpha_1} \cdots x_n^{\alpha_n}$ of degree $d = \sum_{i=1}^n \alpha_i$, the corresponding polynomial of the basis is

\[{d \choose \alpha}^{\frac{1}{2}} x^{\alpha} \quad \text{ where } \quad {d \choose \alpha} = \frac{d!}{\alpha_1! \alpha_2! \cdots \alpha_n!}.\]

For instance, create a polynomial with the basis $[xy^2, xy]$ creates the polynomial $\sqrt{3} a xy^2 + \sqrt{2} b xy$ where a and b are new JuMP decision variables. Constraining the polynomial $axy^2 + bxy$ to be zero with the scaled monomial basis constrains a/√3 and b/√2 to be zero.

[BPT12] Blekherman, G.; Parrilo, P. A. & Thomas, R. R. Semidefinite Optimization and Convex Algebraic Geometry. Society for Industrial and Applied Mathematics, 2012.

PolyJuMP.ZeroConstraint

PolyJuMP.ZeroConstraintWithDomain

PolyJuMP.ZeroPoly

PolyJuMP._warnbounds

PolyJuMP.addpolyconstraint!

PolyJuMP.createpoly

PolyJuMP.cvarchecks

PolyJuMP.eval

PolyJuMP.get_default

PolyJuMP.getdefault

PolyJuMP.getdelegate

PolyJuMP.getdelegates

PolyJuMP.getpolydata

PolyJuMP.getslack

PolyJuMP.include

PolyJuMP.polytype

PolyJuMP.scaling

PolyJuMP.setdefault!

PolyJuMP.setpolymodule!

PolyJuMP.unscale_coef