Docstrings

SimplePosetAlgorithms.SimplePosetAlgorithms

SimplePosetAlgorithms._distinct

SimplePosetAlgorithms.dimension — Function.

dimension(P::SimplePoset, verbose=false) returns the order-theoretic dimension of the poset P. Set verbose to true to see more information as the work is done.

SimplePosetAlgorithms.dimension_work

SimplePosetAlgorithms.eval

SimplePosetAlgorithms.include

SimplePosetAlgorithms.make_linear_order

SimplePosetAlgorithms.max_antichain — Function.

max_antichain(P) returns a largest set of pairwise incomparable elements in the SimplePoset.

SimplePosetAlgorithms.max_chain — Function.

max_chain(P) returns a largest set of pairwise comparable elements in the SimplePoset.

SimplePosetAlgorithms.realize_poset — Function.

realize_poset(R::Array{T,2}) creates a poset from a realizer. The columns of R are the linear extensions of some poset; this function returns that poset.

SimplePosetAlgorithms.realizer — Function.

realizer(P::SimplePoset,d::Int) creates a realizer of P using d linear extensions or throws an error if none exists. The output is an n-by-d matrix whose columns give the linear extensions. The first element in each column is the bottom element of that linear extension.

SimplePosetAlgorithms.width — Function.

width(P) gives the size of a largest antichain in the poset P.