Docstrings

GroupedErrors.@>

GroupedErrors.@across

GroupedErrors.@bootstrap

GroupedErrors.@compare

GroupedErrors.@plot

GroupedErrors.@set_attr

GroupedErrors.@splitby

GroupedErrors.@summarize

GroupedErrors.@x

GroupedErrors.@xlims

GroupedErrors.@xy

GroupedErrors.@y

GroupedErrors.@ylims

GroupedErrors.AbstractSelector

GroupedErrors.ColumnSelector

GroupedErrors.GroupedErrors

GroupedErrors.ProcessedTable

GroupedErrors.Selector

GroupedErrors.Table2Process

GroupedErrors._across

GroupedErrors._bootstrap

GroupedErrors._compare

GroupedErrors._cumulative — Function.

`_cumulative!(df, xaxis, x) = ecdf(df[x])(xaxis)`

Cumulative density function of x, computed along xaxis

GroupedErrors._density — Function.

`_density(df,xaxis::AbstractRange, x; kwargs...)`

Kernel density of x, computed along xaxis

`_density(df, xaxis, x)`

Normalized histogram of x (which is discrete: every value is its own bin)

GroupedErrors._density_axis

GroupedErrors._get

GroupedErrors._get_tup

GroupedErrors._group_apply

GroupedErrors._hazard — Function.

`_hazard(df,xaxis, x; kwargs...)`

Hazard rate of x, computed along xaxis. Keyword arguments are passed to the function computing the density

GroupedErrors._isfinite

GroupedErrors._kw

GroupedErrors._locreg — Function.

`_locreg(df, xaxis::AbstractRange, x, y; kwargs...)`

Apply loess regression, training the regressor with x and y and predicting xaxis

`_locreg(df, xaxis, x, y; estimator = mean)`

In the discrete case, the function computes the estimate of y for a given value of x using the function estimator (default is mean)

GroupedErrors._nafree

GroupedErrors._nafree_tup

GroupedErrors._set_attr

GroupedErrors._splitby

GroupedErrors._summarize

GroupedErrors._x

GroupedErrors._xlims

GroupedErrors._y

GroupedErrors._ylims

GroupedErrors.builtin_axis

GroupedErrors.builtin_funcs

GroupedErrors.eval

GroupedErrors.exampletable

GroupedErrors.functionize

GroupedErrors.get_axis

GroupedErrors.get_grouped_error

GroupedErrors.getcolumn

GroupedErrors.getlength

GroupedErrors.helper_replace_anon_func_syntax

GroupedErrors.include

GroupedErrors.listsplits

GroupedErrors.nsplits

GroupedErrors.pipeline

GroupedErrors.plot_helper

GroupedErrors.process_axis_type!

GroupedErrors.process_function!

GroupedErrors.replace_selector

GroupedErrors.store_kws

GroupedErrors.tablify

GroupedErrors.tuplify