Docstrings

ExprOptimization.@grammar

ExprOptimization.Benchmarks — Module.

Benchmarks for ExprOptimization.jl This will run the julia benchmarks: using ExprOptimization ExprOptimization.Benchmarks.main() See subfolders for the benchmarks of baseline packages Results are placed in RESULTSDIR Call load_csvs() to load up a merged dataframe. See Julia notebook for plots.

ExprOptimization.CrossEntropys.CrossEntropy — Type.

CrossEntropy

Cross Entropy method.

Arguments

pop_size::Int: population size

iterations::Int: number of iterations

max_depth::Int: maximum depth of derivation tree

top_k::Int: top k elite samples used in selection

p_init::Float64: initial value when fitting MLE

init_method::InitializationMethod: Initialization method

ExprOptimization.CrossEntropys

ExprOptimization.ExprOptAlgorithm

ExprOptimization.ExprOptResult — Type.

ExprOptResult

Returned by optimize(). Contains the results of the optimization.

ExprOptimization.ExprOptimization

ExprOptimization.ExprRules

ExprOptimization.ExpressionIterator

ExprOptimization.GeneticPrograms.GeneticProgram — Type.

GeneticProgram

Genetic Programming.

Arguments

pop_size::Int: population size

iterations::Int: number of iterations

max_depth::Int: maximum depth of derivation tree

p_reproduction::Float64: probability of reproduction operator

p_crossover::Float64: probability of crossover operator

p_mutation::Float64: probability of mutation operator

init_method::InitializationMethod: initialization method

select_method::SelectionMethod: selection method

ExprOptimization.GeneticPrograms

ExprOptimization.Grammar

ExprOptimization.GrammaticalEvolutions.GrammaticalEvolution — Type.

GrammaticalEvolution

Grammatical Evolution.

Arguments

grammar::Grammar: grammar

typ::Symbol: start symbol

pop_size::Int: population size

iterations::Int: number of iterations

init_gene_length::Int: initial length of genotype integer array

max_gene_length::Int: maximum length of genotype integer array

max_depth::Int: maximum depth of derivation tree

p_reproduction::Float64: probability of reproduction operator

p_crossover::Float64: probability of crossover operator

p_mutation::Float64: probability of mutation operator

select_method::SelectionMethod: selection method (default: tournament selection)

mutate_method::InitializationMethod: mutation method (default: multi-mutate)

ExprOptimization.GrammaticalEvolutions

ExprOptimization.MonteCarlos.MonteCarlo — Type.

MonteCarlo

Monte Carlo.

Arguments:

num_samples::Int: number of samples

max_depth::Int: maximum depth of derivation tree

ExprOptimization.MonteCarlos

ExprOptimization.NodeLoc

ExprOptimization.NodeRecycler

ExprOptimization.PIPEs.PIPE — Type.

PIPE

Probabilistic Incremental Program Evolution. Example parameters from paper are indicated in parentheses)

Arguments:

ppt_params::PPT: parameters for PPT (e.g., [0.8, 0.2])

pop_size::Int: population size

iterations::Int: number of iterations

p_elitist::Float64: elitist update probability (e.g., 0.2)

c::Float64: learning rate multiplier (e.g., 0.1)

α::Float64: learning rate (e.g., 0.05)

ϵ::Float64: fitness constant (e.g., 1)

p_mutation::Float64: mutation probability (e.g., 0.2)

β::Float64: mutation rate (e.g., 0.6)

p_threshold::Float64: prune threshold (e.g., 0.999)

max_depth::Int: maximum depth of derivation tree

ExprOptimization.PIPEs

ExprOptimization.PPTs.PPT — Type.

PPT

Probabilistic Prototype Tree.

Arguments:

w_terminal::Float64: probability of selecting a terminal on initialization

w_nonterm::Float64: probability of selecting a non-terminal on initialization

ExprOptimization.PPTs

ExprOptimization.ProbabilisticExprRules

ExprOptimization.RuleNode

ExprOptimization.SymbolTable

ExprOptimization.child_types

ExprOptimization.contains_returntype

ExprOptimization.count_expressions

ExprOptimization.depth

ExprOptimization.eval

ExprOptimization.get_executable

ExprOptimization.get_expr — Function.

get_expr(result::ExprOptResult)

Returns the expression in the result

ExprOptimization.include

ExprOptimization.interpret

ExprOptimization.iseval

ExprOptimization.isterminal

ExprOptimization.max_arity

ExprOptimization.mindepth

ExprOptimization.mindepth_map

ExprRules.nchildren — Function.

nchildren(node::PPTNode)

Returns the number of children of a node.

ExprOptimization.node_depth

ExprOptimization.nonterminals

ExprOptimization.optimize — Function.

optimize(p::ExprOptAlgorithm, grammar::Grammar, typ::Symbol, loss::Function; kwargs...)

Main entry for expression optimization. Use concrete ExprOptAlgorithm to specify optimization algorithm. Optimize using grammar and start symbol, typ, and loss function. Loss function has the form: los::Float64=loss(node::RuleNode).

ExprOptimization.recycle!

ExprOptimization.return_type

ExprOptimization.root_node_loc

ExprOptimization.sample