Docstrings

Poltergeist...

Poltergeist.AR_process — Function.

AR_process(f,A::Fun)

Calculate coefficients of Gaussian AR process with the same autocorrelation function as A under the map f.

See Appendix A.2 of Wormell, C.L. & Gottwald, G.A. 'On the Validity of Linear Response Theory in High-Dimensional Deterministic Dynamical Systems' (2018).

Poltergeist.AbstractBranch

Poltergeist.AbstractCircleMap

Poltergeist.AbstractIntervalMap

Poltergeist.AbstractMarkovMap

Poltergeist.AbstractTransfer

Poltergeist.BackSum

Poltergeist.BackSumEval

Poltergeist.BasisFun

Poltergeist.BasisFunInt

Poltergeist.CircleMap

Poltergeist.ComposedCircleMap

Poltergeist.ComposedMap

Poltergeist.ComposedMarkovMap

Poltergeist.ConcreteTransfer

Poltergeist.Direction

Poltergeist.ExpandingBranch

Poltergeist.ExpandingBranchDerivative

Poltergeist.ExpandingBranchDerivativeInverse

Poltergeist.ExpandingBranchInverse

Poltergeist.Forward

Poltergeist.ForwardDirection

Poltergeist.Fun

Poltergeist.FunctionDerivative

Poltergeist.FwdCircleMap

Poltergeist.FwdExpandingBranch

Poltergeist.FwdOffset

Poltergeist.GeneralInterval

Poltergeist.HGraphs

Poltergeist.HofbauerDomain — Type.

HofbauerDomain

A Domain object annotated with a depth index for HofbauerExtensions. It is not a Domain subtype itself as it should only be used in the context of Hofbauer extensions.

Poltergeist.HofbauerPoint — Type.

HofbauerPoint(x, graphind)

Constructs a HofbauerPoint, representing point x on the graphindth domain of a Hofbauer extension.

Poltergeist.InducedMap

Poltergeist.InterpolationNode

Poltergeist.Interval

Poltergeist.IntervalMap

Poltergeist.LyapContainer

Poltergeist.MA_process — Function.

MA_process(f,A::Fun)

Calculate coefficients of Gaussian MA process with the same autocorrelation function as A under the map f.

See Appendix A.2 of Wormell, C.L. & Gottwald, G.A. 'On the Validity of Linear Response Theory in High-Dimensional Deterministic Dynamical Systems' (2018).

Poltergeist.MarkovMap — Type.

MarkovMap(branches::Vector, domain, rangedomain)

Generate a computer representation of a full-branch uniformly-expanding interval map domain → rangedomain using a vector describing the branches of the map.

Poltergeist.MarkovMapDerivative

Poltergeist.Matrix

Poltergeist.NeutralBranch

Poltergeist.PeriodicSegment

Poltergeist.Poltergeist

Poltergeist.PureLaurent

Poltergeist.RevCircleMap

Poltergeist.RevExpandingBranch

Poltergeist.RevOffset

Poltergeist.Reverse

Poltergeist.ReverseDirection

Poltergeist.SimpleBranchedMap

Poltergeist.SolutionInv

Poltergeist.SolutionInvWrapper

Poltergeist.Transfer — Function.

Transfer(m::AbstractMarkovMap)

Create a CachedOperator of a ConcreteTransfer operator encoding the transfer operator of m.

Caching is used for speed, as entries of the transfer operator are most efficiently calculated whole columns at a time.

Poltergeist.TransferBSFunction

Poltergeist.TransferCall

Poltergeist.TransferIntBSFunction

Poltergeist._NI

Poltergeist._hofbauerdestination

Poltergeist._hofbauersource

Poltergeist.acim — Function.

acim(L)

Output a Fun object giving the acim of the associated map

Poltergeist.approx_in

Poltergeist.approx_issubset

Poltergeist.atol_isapprox

Poltergeist.autodiff

Poltergeist.autodiff_dual

Poltergeist.backsumrecursion

Poltergeist.basic_newton

Poltergeist.birkhoffcov — Function.

birkhoffcov(L, A::Fun, B::Fun)

Output covariance of CLT-normalised Birkhoff sums of A and B under the map

Poltergeist.birkhoffvar — Function.

birkhoffvar(L, A::Fun)

Output diffusion coefficient of A under the map using Green-Kubo formula

Poltergeist.branch

Poltergeist.branches — Function.

branches(m)

Return the branches of the map m.

Poltergeist.branchindtype

Poltergeist.chebyTk

Poltergeist.chebyTk_int

Poltergeist.complength

Poltergeist.containsnfp

Poltergeist.containstransfer

Poltergeist.correlationsum — Function.

correlationsum(L, A::Fun, ρ=acim(L))

Output resolvent of transfer operator applied to ρ*(A - ∫A dρ)

Poltergeist.covariancefunction — Function.

cfA = covariancefunction(L, A, n)

Compute the lag-covariance function against transfer operator L for observable A for n time steps in each direction. Specifically, the (k+1)th entry is the expectation of A A∘f^(k).

cf = covariancefunction(L, A; tol = eps(A))

Choose n so that the covariance declines to a given tolerance

cfA, cfB = covariancefunction(L, A, B, n)

Compute the lag-covariance function against transfer operator L between observables A and B for n time steps in each direction. Specifically, the (k+1)th entry of cfA is the expectation of B A∘f^(k) and vice versa for cfB.

cfA, cfB = covariancefunction(L, A, B; tol=eps(A*B))

Chooses n so that the covariance declines to a given tolerance.

Poltergeist.coveringinterval

Poltergeist.cumsum

Poltergeist.disc_newton

Poltergeist.domain_guess

Poltergeist.domain_mod

Poltergeist.domain_newton

Poltergeist.doubling — Function.

doubling(d = Interval(0,1.))

Returns the full-branch interval map on domain d with 2 equally-sized branches.

See also: tupling, lanford

Poltergeist.eachbranchindex — Function.

eachbranchindex(m)

Return an iterator giving the indices of the branches of m

Poltergeist.eigs

Poltergeist.eval

Poltergeist.forwardmodulomap

Poltergeist.fourierCSk

Poltergeist.fourierCSk_int

Poltergeist.getbasisfun

Poltergeist.getbasisfun_int

Poltergeist.getbranchind

Poltergeist.haswrapperstructure

Poltergeist.hofbauerextension — Function.

hofbauerextension(m,basedomains;maxdepth=100,forcereturn=trues(length(basedomains)))

Generates a Hofbauer extension of m starting from basedomains, which may be a single domain.

The keyword maxdepth says how deep the Hofbauer extension should go, and forcereturn sets whether a return to given members of basedomains should forced if possible.

For example, in the case of f(x) = mod(2x,1), if basedomains is set to Interval(0.,0.3) then Interval(0.,0.5) might map to Interval(0.,1.)for a given map ifforcereturn=falsebut would map to Interval(0,0.3) and Interval(0.3,1) if forcereturn=true.

Poltergeist.include

Poltergeist.induce

Poltergeist.interval_guess

Poltergeist.interval_mod

Poltergeist.interval_newton

Poltergeist.lanford — Function.

lanford(T=Float64)

Return the Lanford map, with type encoding T.

See also: tupling, doubling

Poltergeist.lanford_brk

Poltergeist.lanford_dv0

Poltergeist.lanford_dv1

Poltergeist.lanford_v0

Poltergeist.lanford_v1

Poltergeist.linearresponse — Function.

linearresponse(L, X::Fun)

Output a Fun object giving the first-order change in the acim of the map under perturbation X

Poltergeist.logMA_process

Poltergeist.lyapunov — Function.

lyapunov(f, r=acim(f), sp=Space(rangedomain(f)))

Calculate Lyapunov exponent associated with f

Poltergeist.mapD

Poltergeist.mapL

Poltergeist.mapP

Poltergeist.map_n

Poltergeist.maphb — Function.

maphb(b::AbstractBranch, x)

Maps point x according to b. For NeutralBranches calculates the return map out of domain(b).

Poltergeist.maphbD

Poltergeist.maphbP

Poltergeist.maphbinv

Poltergeist.maphbinvD

Poltergeist.maphbinvP

Poltergeist.mapinterval

Poltergeist.mapinv

Poltergeist.mapinvD

Poltergeist.mapinvP

Poltergeist.markovmap

Poltergeist.maxrangedomain

Poltergeist.modulomap — Function.

modulomap(f, D, R=dom; diff= autodiff(f,dom))

Output MarkovMap or CircleMap m: D → R such that m(x) = f(x) mod R.

Poltergeist.nbranches — Function.

nbranches(m)

Return the number of branches of m.

Poltergeist.ncover

Poltergeist.neutralfixedpoints

Poltergeist.nneutral

Poltergeist.perturb — Function.

perturb(d, X, ϵ)

Construct a self-map on domain d: x ↦ x + ϵ X(x)

perturb(m::AbstractMarkovMap, X, ϵ)

Output perturbation of m: x ↦ m(x) + ϵ X(m(x))

Poltergeist.rangedomain

Poltergeist.resizecolstops!

Poltergeist.reversemodulomap

Poltergeist.samplable

Poltergeist.setcolstops!

Poltergeist.show

Poltergeist.timehist

Poltergeist.timehist_initialvalue

Poltergeist.timeseries

Poltergeist.timeseries_initialvalue

Poltergeist.towerextend!

Poltergeist.towerwalk!

Poltergeist.transfer

Poltergeist.transfer_getindex

Poltergeist.transferbranch

Poltergeist.transferbranch_int

Poltergeist.transferbranch_int_edges

Poltergeist.transferfunction

Poltergeist.transferfunction_int

Poltergeist.transferfunction_nodes

Poltergeist.tupling — Function.

tupling(k::Int, d = Interval(0,1.))

Returns the full-branch interval map on domain d with k equally-sized branches.

See also: doubling, lanford

Poltergeist.uniform

Poltergeist.unsafe_call

Poltergeist.unsafe_inv

Poltergeist.unsafe_mapD

Poltergeist.unsafe_mapP

Poltergeist.unsafe_mapinv

Poltergeist.unsafe_mapinvD

Poltergeist.unsafe_mapinvP

Poltergeist.zero_to — Function.

zero_to(A::Fun, ρ::Fun=uniform(space(A)))

Output ρ*(A - ∫A dρ)