Docstrings

WiltonInts84.CLOCKWISE

WiltonInts84.COUNTERCLOCKWISE

WiltonInts84.IntegrationPath

WiltonInts84.WiltonInts84

WiltonInts84.WorkSpace

WiltonInts84.add

WiltonInts84.angle — Function.

angle(p,q)

Returns the positive angle in [0,π] between p and q

WiltonInts84.arcintsg

WiltonInts84.buildgrad

WiltonInts84.circleintsg

WiltonInts84.contour

WiltonInts84.contour!

WiltonInts84.distancetoline

WiltonInts84.eval

WiltonInts84.incidenceLineWithSphere — Function.

r, t = incidenceLineWithSphere(v, first, last, r)

Returns the number of crossings and their barycentric coordinates for a circle with center v and radius r and a segment [a,b]. If the circle is tangent to the segment, zero crossings are reported.

WiltonInts84.include

WiltonInts84.inside

WiltonInts84.maketuple

WiltonInts84.segintsg

WiltonInts84.wiltonints — Function.

wiltonints(p1,p2,p3,x,[r,R],Val{N})

Compute potential integrals over a triangle (intersected with a spherical) mesh. Powers of the distance up to degree N are computed.

WiltonInts84.workspace