Docstrings

SimpleDrawing.Cubic — Type.

Cubic(a,b,c,d) is a 3rd degree polynomial a+bx+cx^2+dx^3. Use f(x) to evalue f at the value x.

SimpleDrawing.SimpleDrawing

SimpleDrawing.Spline — Type.

Spline(vals,kind) returns a cubic spline based on the values in vals. The resulting spline S will have the property that S(1)==y[1], S(2)==y[2], and so on up to S(n)==y[n] where n is the length of y.

If kind is :open then the second derivatives at the end points will be zero. (This is the default.)

If kind is :closed then we assume that we are interpolating a periodic function where S(n+1)==S(1).

SimpleDrawing.closed_spline

SimpleDrawing.draw

SimpleDrawing.draw_arc — Function.

draw_arc(x,y,r,t1,t2) draws the shorter arc centered at (x,y) with radius r between angles t1 and t2.

draw_arc(a,b,c) where the arguments are Complex: Draw the arc from a through b to c.

SimpleDrawing.draw_circle — Function.

draw_circle(x,y,r) (or draw_circle(z,r)) draws a circle centered at (x,y) (at z) with radius r.

draw_circle(a,b,c) draws a circle through the points given by the three complex arguments.

SimpleDrawing.draw_curve — Function.

draw_curve(plist) draws a closed curve through points in the plist which is a 1-dimensional list of complex numbers. Use draw_curve(plist,false) for an open curve.

SimpleDrawing.draw_point — Function.

draw_point(x,y) [or draw_point(z)] draws a point at coordinates (x,y) [or complex location z].

draw_point(list) plots all the points in the one-dimensional list of complex values.

SimpleDrawing.draw_segment — Function.

draw_segment(x,y,xx,yy) draws a line segment from (x,y) to (xx,yy). Also draw_segment(z,zz) for Complex arguments.

SimpleDrawing.draw_vector — Function.

draw_vector is used to draw vectors (line segments with an arrow at one end). The variations are:

draw_vector(x,y) draws a vector from the origin to (x,y).

draw_vector(x,y,basex,basey) draws a vector from (basex,basey) to

(basex+x,basey+y).

draw_vector(z) draws a vector from the origin to the complex value z.

draw_vector(z,basez) draws a vector from the complex location basez to

z+basez.

SimpleDrawing.eval

SimpleDrawing.find_center — Function.

find_center(a,b,c): Given three points in the complex plane, find the point z that is equidistant from all three. If the three points are collinear then return Inf + Inf*im.

SimpleDrawing.finish — Function.

finish() is used to clean up a drawing after various calls to draw. It removes the axes and the grid lines, and sets the aspect ratio to one.

SimpleDrawing.funk — Function.

funk(S) converts the Spline S into a callable function (e.g., that can be passed to plot).

SimpleDrawing.include

SimpleDrawing.is_closed

SimpleDrawing.newdraw — Function.

newdraw() clears the screen and sets up a blank drawing canvas with no axis, grid, or legend, and sets the aspect ratio to 1.

SimpleDrawing.non_colinear_check — Function.

non_colinear_check(a,b,c): test if the complex numbers are distinct and noncollinear.

SimpleDrawing.npatches

SimpleDrawing.open_spline