Docstrings

PolynomialFactors.AbstractFactorTree

PolynomialFactors.FactorTree

PolynomialFactors.FactorTree_over_Zp

PolynomialFactors.LLL_basis_reduction!

PolynomialFactors.PolynomialFactors

PolynomialFactors._equal_degree_splitting

PolynomialFactors._factor_combinations

PolynomialFactors._modp

PolynomialFactors._powermod

PolynomialFactors._reduce

PolynomialFactors.all_children

PolynomialFactors.as_poly

PolynomialFactors.as_poly_Zp

PolynomialFactors.as_poly_modp

PolynomialFactors.deflate

PolynomialFactors.distinct_degree_factorization

PolynomialFactors.equal_degree_factorization

PolynomialFactors.equal_degree_splitting

PolynomialFactors.eval

PolynomialFactors.exact_divrem — Function.

Division algorithm for Z[x]

returns q,r with

a = q*b + r

degree(r) < degree(b)

If no such q and r can be found, then both q and r are 0.

PolynomialFactors.factor_Zp

PolynomialFactors.factor_Zp_squarefree

PolynomialFactors.factor_Zx_big_prime_squarefree

PolynomialFactors.factor_Zx_prime_power_squarefree

PolynomialFactors.factormod

PolynomialFactors.has_children

PolynomialFactors.hensel_lift — Function.

Algo 15.17 multifactor Hensel lifting

PolynomialFactors.hensel_step

PolynomialFactors.hensel_step_update_factor_tree!

PolynomialFactors.identify_factor

PolynomialFactors.identify_factors_lll — Function.

Use LLL basis reduction to identify factors of square free f from its factors over Zp

PolynomialFactors.include

PolynomialFactors.make_factor_tree_over_Zp

PolynomialFactors.maxnorm

PolynomialFactors.mu

PolynomialFactors.onenorm

PolynomialFactors.poly_coeffs

PolynomialFactors.poly_factor

PolynomialFactors.short_vector

PolynomialFactors.square_free

PolynomialFactors.swaprows!

PolynomialFactors.tuplejoin

PolynomialFactors.value

PolynomialFactors.variable