Docstrings

TypedPolynomials.@polyvar — Macro.

Construct polynomial variables and bind them to local variables of the same name. Usage:

Create a single variable named x:

@polyvar(x)

Create several variables at the same time:

@polyvar(x, y, z)

Create a tuple of variables x = (x1, x2, x3, x4):

@polyvar(x[1:4])

Create variables x, y, a = (a1, a2, a3):

@polyvar(x, y, a[1:3])

You can also assign the results of the macro to a tuple:

vars = @polyvar(x[1:5])
@assert typeof(vars[1]) == Variable{:x1}

TypedPolynomials.MP

TypedPolynomials.Monomial

TypedPolynomials.MonomialLike

TypedPolynomials.Polynomial

TypedPolynomials.PolynomialLike

TypedPolynomials.Sequences

TypedPolynomials.Substitution

TypedPolynomials.Term

TypedPolynomials.TermLike

TypedPolynomials.TypedPolynomials

TypedPolynomials.Variable

TypedPolynomials._diff

TypedPolynomials._exponent

TypedPolynomials._makevar

TypedPolynomials._promote_monomial_noncommutative

TypedPolynomials._return_name

TypedPolynomials._split

TypedPolynomials._varconstructor

TypedPolynomials.changeeltype

TypedPolynomials.checksorted

TypedPolynomials.coefftype

TypedPolynomials.combine

TypedPolynomials.compare

TypedPolynomials.differentiate

TypedPolynomials.eval

TypedPolynomials.find_variable_index

TypedPolynomials.include

TypedPolynomials.inmonomial

TypedPolynomials.jointerms

TypedPolynomials.matchindices

TypedPolynomials.merge

TypedPolynomials.monomial_powers

TypedPolynomials.subs

TypedPolynomials.terms

TypedPolynomials.variables