Docstrings

ParallelProcessingTools.@critical — Macro.

@critical expr

Mark code in expr as a critical section. Code in critical sections will never be executed in parallel (via multithreading) to any other critical section.

@critical is very useful to mark non-threadsafe code.

Example:

@onthreads allthreads() begin
    @critical @info Base.Threads.threadid()
end

Without `@critical`, the above will typically crash Julia.

ParallelProcessingTools.@onallthreads

ParallelProcessingTools.@onprocs — Macro.

@onprocs procsel expr

Executes expr in parallel on all processes in procsel. Waits until all processes are done. Returns all results as a vector (or as a single scalar value, if procsel itself is a scalar).

Example:

using Distributed
addprocs(2)
workers() == @onprocs workers() myid()

ParallelProcessingTools.@onthreads — Macro.

@onthreads threadsel expr

Execute code in expr in parallel on the threads in threadsel.

threadsel should be a single thread-ID or a range (or array) of thread-ids. If threadsel == Base.Threads.threadid(), expr is run on the current tread with only minimal overhead.

Note: Currently, multiple @onthreads sections will not run in parallel to each other, even if they use disjunct sets of threads, due to limitations of the Julia multithreading implementation. This restriction is likely to disappear in future Julia versions.

In contrast to Base.Threads.@threads, @onthreads does forward exceptions to the caller.

Example 1:

tlsum = ThreadLocal(0.0)
data = rand(100)
@onthreads allthreads() begin
    tlsum[] = sum(workpart(data, allthreads(), Base.Threads.threadid()))
end
sum(getallvalues(tlsum)) ≈ sum(data)

Example 2:

# Assuming 4 threads:
tl = ThreadLocal(42)
threadsel = 2:3
@onthreads threadsel begin
    tl[] = Base.Threads.threadid()
end
getallvalues(tl)[threadsel] == [2, 3]
getallvalues(tl)[[1,4]] == [42, 42]

ParallelProcessingTools.AbstractThreadLocal — Type.

abstract type AbstractThreadLocal{T} end

Abstract type for thread-local values of type T.

The value for the current thread is accessed via getindex(::AbstractThreadLocal) and `setindex(::AbstractThreadLocal, x).

To access both regular and thread-local values in a unified manner, use the function getlocalvalue.

To get the all values across all threads, use the function getallvalues.

Default implementation is ThreadLocal.

ParallelProcessingTools.LockableIO

ParallelProcessingTools.LockableValue

ParallelProcessingTools.ParallelProcessingTools

ParallelProcessingTools.ThreadLocal — Type.

ThreadLocal{T} <: AbstractThreadLocal{T}

Represents a thread-local value. See AbstractThreadLocal for the API.

Constructors:

ThreadLocal{T}() where {T}
ThreadLocal(value::T) where {T}
ThreadLocal{T}(f::Base.Callable) where {T}

Examples:

tlvalue = ThreadLocal(0)
@onthreads allthreads() tlvalue[] = Base.Threads.threadid()
getallvalues(tlvalue) == allthreads()

rand_value_on_each_thread = ThreadLocal{Float64}(rand)
all(x -> 0 < x < 1, getallvalues(rand_value_on_each_thread))

ParallelProcessingTools.ThreadSafe

ParallelProcessingTools.ThreadSafeIO

ParallelProcessingTools.ThreadSafeReentrantLock

ParallelProcessingTools._check_thread_local_errors

ParallelProcessingTools._check_threadsel

ParallelProcessingTools._clear_thread_local_errors

ParallelProcessingTools._critical_section_lock

ParallelProcessingTools._current_thread_selected

ParallelProcessingTools._is_sorted_and_unique

ParallelProcessingTools._protect_from_resize

ParallelProcessingTools._run_on_procs

ParallelProcessingTools._run_on_threads

ParallelProcessingTools._set_thread_local_error

ParallelProcessingTools._thread_exec_func

ParallelProcessingTools._thread_local_error_err

ParallelProcessingTools._thread_local_error_set

ParallelProcessingTools._workpart_hi

ParallelProcessingTools._workpart_scheme

ParallelProcessingTools.allthreads — Function.

allthreads()

Convencience function, returns 1:Base.Threads.nthreads().

ParallelProcessingTools.eval

ParallelProcessingTools.getallvalues — Function.

getallvalues(v::AbstractThreadLocal{T})::AbstractVector{T}

Access the all values (one for each thread) of a thread-local value as a vector. Can only be called in single-threaded code sections.

ParallelProcessingTools.getlocalvalue — Function.

getlocalvalue(x::Any) = x
getlocalvalue(x::ThreadLocal) = x[]

Access plain values and thread-local values in a unified fashion.

ParallelProcessingTools.include

ParallelProcessingTools.isdefined_local

ParallelProcessingTools.mtjulia_exe

ParallelProcessingTools.threadpartition

ParallelProcessingTools.workpart — Function.

workpart(data::AbstractArray, workersel::AbstractVector{W}, current_worker::W) where {W}

Get the part of data that the execution unit current_worker is responsible for. Implies a partition of data across the workers listed in workersel.

For generic data arrays, workpart will return a view. If data is a Range (e.g. indices to be processed), a sub-range will be returned.

Type W will typically be Int and workersel will usually be a range/array of thread/process IDs.

Note: workersel is required to be sorted in ascending order and to contain no duplicate entries.

Examples:

using Distributed, Base.Threads
A = rand(100)
# ...
sub_A = workpart(A, workers(), myid())
# ...
idxs = workpart(eachindex(sub_A), allthreads(), threadid())
for i in idxs
    # ...
end

ParallelProcessingTools.workpartition