Docstrings

WorldBankData.WorldBankData — Module.

Provides two functions, search_wdi and wdi, for searching and fetching World Development Indicators data from the World Bank.

WorldBankData.all_countries

WorldBankData.clean_append!

WorldBankData.clean_entry

WorldBankData.convert_a2f

WorldBankData.country_cache

WorldBankData.country_match

WorldBankData.df_match

WorldBankData.download_countries

WorldBankData.download_indicators

WorldBankData.download_parse_json

WorldBankData.eval

WorldBankData.get_countries

WorldBankData.get_indicators

WorldBankData.include

WorldBankData.indicator_cache

WorldBankData.indicator_match

WorldBankData.make_symbol

WorldBankData.parse_country

WorldBankData.parse_indicator

WorldBankData.parse_wdi

WorldBankData.regex_match

WorldBankData.reset_country_cache

WorldBankData.reset_indicator_cache

WorldBankData.search_countries

WorldBankData.search_indicators

WorldBankData.search_wdi — Function.

search_wdi(data::String, entry::String, regx::Regex)::DataFrame

Search World Development Indicators for countries or indicators.

https://datacatalog.worldbank.org/dataset/world-development-indicators

Arguments

data : data to search for: "indicators" or "countries"

entry : entry to lookup for countries: name,region,capital,iso2c,iso3c,income,lending for indicators: name, description, topics, source_database, source_organization

regex : regular expression to find

Examples

search_wdi("countries", "name", r"united"i)
search_wdi("indicators", "description", r"gross national"i)

WorldBankData.set_country_cache

WorldBankData.set_indicator_cache

WorldBankData.tofloat

WorldBankData.wdi — Function.

wdi(indicators::Union{String,Array{String,1}}, countries::Union{String,Array{String,1}}, startyear::Integer=1800, endyear::Integer=3000; extra::Bool=false, verbose::Bool=false)::DataFrame

Download data from World Development Indicators (WDI) Data Catalog of the World Bank.

https://datacatalog.worldbank.org/dataset/world-development-indicators

Arguments indicators : indicator name or array of indicators countries : string or string array of ISO 2 letter country codes or all for all countries. startyear : first year to include endyear : last year to include extra : if true additional country data should be included (region, capital, longitude, latitude, income, lending) verbose : if true print URLs downloaded, useful as progress indicator.

Examples

df = wdi("NY.GNP.PCAP.CD", "US", 1980, 2012, extra=true)
df = wdi("NY.GNP.PCAP.CD", ["US","BR"], 1980, 2012, extra=true)
df = wdi(["NY.GNP.PCAP.CD", "AG.LND.ARBL.HA.PC"], ["US","BR"], 1980, 2012, extra=true)

WorldBankData.wdi_download