Docstrings

PolynomialZeros.AA

PolynomialZeros.AGCD

PolynomialZeros.AMVW

PolynomialZeros.MultRoot

PolynomialZeros.Over — Module.

Use types in this module to specifiy a domain.

PolynomialZeros.PolynomialZeros

PolynomialZeros.RealRoots

PolynomialZeros._rational_T

PolynomialZeros.as_poly — Function.

`as_poly(f)` convert something into a polynomial

Here f can be a Polynomial object; a vector of coefficents in the form [a_0, a_1, ..., a_n]; or a callable object, such as a function, that implements a polynomial function. The latter is determined by whether it can be evaluated on the Polynomial monomial x.

as_poly{T}(T, f)

Convert f to a polynomial of type Poly{T}, or throw a DomainError.

PolynomialZeros.bisection

PolynomialZeros.check_cubic

PolynomialZeros.check_linear

PolynomialZeros.check_quadratic

PolynomialZeros.check_quintic

PolynomialZeros.discr

PolynomialZeros.e_type — Function.

Type of polynomial

PolynomialZeros.eval

PolynomialZeros.include

PolynomialZeros.monic — Function.

monic

PolynomialZeros.poly_coeffs — Function.

Find coefficients of polynomial expressed as Poly, Callable object, or values [a0,a1, ..., an]

PolynomialZeros.poly_roots — Function.

poly_roots(f, domain): Find zeros of the polynomial f within the specified domain.

f can be an instance of Poly type (from Polynomials.jl) or a callable object which can be so converted. Will throw a DomainError if the object f can not be converted to Poly{T}.

domain is one of

over.C (the default) for solving over complex values (Complex{Float64}). Use Over.CC{T} to specfy a type T<: AbstractFloat other than Float64. The default method is from PolynomialRoots. Pass the argument method=:roots to use the roots function from Polynomials.jl. Pass the argument method=:amvw to use an algorithm by Aurentz, Mach, Vandebril, and Watkins. For a degree n polynomial over C, all n roots should be returned (including multiplicities).

over.R for solving over the real line (Float64). Use Over.RR{T} to specify a T <: Integer other than Float64. The algorithm assumes the polynomial is square free (none of its factors are squares over R). This is important for floating point coefficients. Pass the argument square_free=false to have an approximate gcd used to create a square-free version. Only unique real roots are returned (no multiplicities).

over.Q for solving over the rational numbers (Rational{Int}). Use Over.RR{T} to specify a T <: Integer other than Int. Only unique rational roots are returned (no multiplicities).

over.Z for solving over the integers (Int). Use Over.ZZ{T} to specify a T other than Int. Only unique integer roots are returned (no multiplicities).

over.Zp{p} for solving over the finite field ZZ_p, p a prime. Only unique roots are returned (no multiplicities).

Returns an array of zeros, possibly empty. May throw error if polynomial type is inappropriate for specified domain. For an unspecified domain, the domain may reflect the element type of f(0).

Examples:

using Polynomials, PolynomialZeros
x = variable(); p = x^5 - x - 1
poly_roots(p, Over.C)  # 5
poly_roots(p, Over.R)  # 1
poly_roots(p, Over.Q)  # empty

p = x^3 - 1
poly_roots(p, Over.C)  # 3
poly_roots(p, Over.R)  # 1
poly_roots(p, Over.Q)  # 1
poly_roots(p, Over.Z)  # 1
poly_roots(p, Over.Zp{7})  # 3

PolynomialZeros.qdrtc

PolynomialZeros.rcoeffs — Function.

reverse coefficients of a polynomial

PolynomialZeros.solve_cubic

PolynomialZeros.solve_linear

PolynomialZeros.solve_quadratic

PolynomialZeros.solve_quintic

PolynomialZeros.special_case