Docstrings

TeaSeis.DataProperty

TeaSeis.Extent

TeaSeis.Geometry — Type.

g = Geometry(;ox=0.0,oy=0.0,oz=0.0,ux=1.0,uy=0.0,uz=0.0,vx=0.0,vy=1.0,vz=0.0,wx=0.0,wy=0.0,wz=1.0,u1=0,un=0,v1=0,vn=0,w1=0,wn=0)

where g::Geometry. The named arguments are:

ox=0.0,oy=0.0,oz=0.0 origin of axes

ux=1.0,uy=0.0,uz=0.0 end of u-vector (e.g. end of first in-line, in the cross-line direction

vx=0.0,vy=1.0,vz=0.0 end of v-vector (e.g. end of first cross-line, in the in-line direction

wx=0.0,wy=0.0,wz=1.0 end of depth axis

u1=1 minimum index along the u-vector (e.g. maximum cross-line index)

un=2 maximum index along the u-vector (e.g. maximum cross-line index)

v1=1 minimum index along the v-vector (e.g. minimum in-line index)

vn=2 maximum index along the v-vector (e.g. maximum in-line index)

w1=1 minimum depth index

wn=2 maximum depth index

TeaSeis.JSeis

TeaSeis.TeaSeis

TeaSeis.TraceCompressor

TeaSeis.TraceProperty

TeaSeis.TracePropertyDef

TeaSeis.allocframe — Function.

allocframe(io)

Allocate memory for one frame of JavaSeis dataset. Returns (Array{Float32,2},Array{UInt8,2}). For example, trcs, hdrs = allocframe(jsopen("data.js")).

TeaSeis.allocframebuf

TeaSeis.allocframehdrs — Function.

allocframehdrs(io)

Allocate memory for headers for one frame of JavaSeis dataset. Returns Array{UInt8,2}. For example, hdrs = allocframehdrs(jsopen("data.js")).

TeaSeis.allocframetrcs — Function.

allocframetrcs(io)

Allocate memory for traces for one frame of JavaSeis dataset. Returns Array{Float32,2}. For example, trcs = allocframetrcs(jsopen("data.js")).

TeaSeis.asciidataformat

TeaSeis.create_map

TeaSeis.dataproperty — Function.

dataproperty(io, label)

Get a data property (data properties are per file, rather than per trace) from io::JSeis with label label::String. For example, dataproperty(jsopen("data.js"), "FREQUENCY").

TeaSeis.datestamp

TeaSeis.delete_first_line

TeaSeis.description

TeaSeis.dictJStoPM

TeaSeis.dictPMtoJS

TeaSeis.domains — Function.

domains(io)

Returns the domains of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

domains(io, i)

Returns the domain of the ith dimension of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.elementcount

TeaSeis.emptysecondaries!

TeaSeis.eval

TeaSeis.extentdir

TeaSeis.extentindex

TeaSeis.fold — Function.

fold(io, hdrs)

Compute the fold of a frame where io is JSeis corresponding to the dataset, and hdrs are the headers for the frame. For example: io=jsopen("file.js"); fold(io, readframehdrs(io,1))

fold(io, idx...)

Compute the fold of a frame where idx is the frame/volume/hypercube indices. For example, fold(jsopen("file.js"),1) for a 3D dataset, fold(jsopen("file.js",1,2)) for a 4D dataset, and fold(jsopen("file.js"),1,2,3) for a 5D dataset.

TeaSeis.fold!

TeaSeis.fold_impl

TeaSeis.format

TeaSeis.formataxes

TeaSeis.geometry — Function.

geometry(io)

If io::JSeis contains a geometry definition, then return a geometry of type Geometry. Otherwise, return nothing.

TeaSeis.get_axis_domains

TeaSeis.get_axis_lengths

TeaSeis.get_axis_lincs

TeaSeis.get_axis_lstarts

TeaSeis.get_axis_pincs

TeaSeis.get_axis_propdef

TeaSeis.get_axis_propdefs

TeaSeis.get_axis_pstarts

TeaSeis.get_axis_units

TeaSeis.get_dataformat

TeaSeis.get_dataorder

TeaSeis.get_dataproperties

TeaSeis.get_datatype

TeaSeis.get_description

TeaSeis.get_extent_manager_element

TeaSeis.get_extentmaxpos

TeaSeis.get_extentname

TeaSeis.get_extents

TeaSeis.get_extentsize

TeaSeis.get_file_property_element

TeaSeis.get_geom

TeaSeis.get_mapped

TeaSeis.get_nextents

TeaSeis.get_nsecondaries

TeaSeis.get_secondaries

TeaSeis.get_secondary

TeaSeis.get_status

TeaSeis.get_trace_properties

TeaSeis.hasdataproperty — Function.

hasdataproperty(io, label)

return true if io::JSeis contains the data property corresponding to label. Otherwise, return false.

TeaSeis.headerlength

TeaSeis.include

TeaSeis.ind2sub — Function.

ind2sub(io, i)

Return the (frame,volume...) tuple for the liner index i. This is useful for looping over all frames in a data-set that is more that 4 or more dimensions. For example,

for i = 1:length(io)
    trcs, hdrs = readframe(io, ind2sub(io,i)...)
end

TeaSeis.index

TeaSeis.jscreate — Function.

jscreate(filename)

Create a JavaSeis dataset without opening it. This method has the same optional arguments as jsopen

TeaSeis.jsopen — Function.

jsopen(filename, mode, [parameters])

Open a new or existing JavaSeis dataset with name filename::String and in mode::String. mode can be one of "r" (read), "w" (write/create) or "r+" (read and write). It is convention for filename to havea ".js" extention.

If "w" is used for the value of mode, then the axis_lengths named parameter is required, and several optional named function parameters are available:

parameters

similarto::String An existing JavaSeis dataset. If set, then all other named arguments can be used to modify the data context that belongs to the existing JavaSeis dataset.

description::String Description of dataset, if not set, then a description is parsed from the filename.

mapped::Bool If the dataset is full (no missing frames/traces), then it may be more efficient to set this to false. Defaults to true.

datatype::String Examples are CMP, SHOT, etc. If not set, then UNKNOWN is used.

dataformat::Type Choose from Float32, and Int16. If not set, then Float32 is used.

dataorder::String (not supported)

axis_lengths::Array{Int} size of each dimension (sample/trace/frame/volume/hypercube) of the JavaSeis data context

axis_propdefs::Array{TracePropertyDef} Trace properties corresponding to JavaSeis axes. If not set, then SAMPLE, TRACE, FRAME, VOLUME and HYPRCUBE are used.

axis_units::Array{String} Units corresponding to JavaSeis axes. e.g. SECONDS, METERS, etc. If not set, then UNKNOWN is used.

axis_domains::Array{String} Domains corresponding to JavaSeis axes. e.g. SPACE, TIME, etc. If not set, then UNKNOWN is used.

axis_lstarts::Array{Int} Logical origins for each axis. If not set, then 1 is used for the logical origin of each axis.

axis_lincs::Array{Int} Logical increments for each axis. If not set, then 1 is used for the logical increments of each axis.

axis_pstarts::Array{Float64} Physical origins for each axis. If not set, then 0.0 is used for the physical origin of each axis.

axis_pincs::Array{Float64} Physical increments for each axis. If not set, then 1.0 is used for the physical increments of each axis.

data_properties::Array{DataProperty} An array of custom trace properties. These are in addition to the properties listed in SSPROPS.md.

properties::Array{TracePropertyDef} An array of custom data properties. One property per data-set rather than one property per trace as in properties above.

geometry::Geometry An optional three point geometry can be embedded in the JavaSeis file.

secondaries::Array{String} An array of file-system locations used to store the file extents. If not set, then primary storage is used.

nextents::Int64 The number of file-extents used to store the data. If not set, then a heuristic is used to choose the number of extents. The heuristic is: min(256,10 + (FRAMEWORK_SIZE)/(2*1024^3)).

properties_add::Array{TracePropertyDef} When similarto is specified, use this to add trace properties to those already existing in the similarto file.

properties_rm::Array{TracePropertyDef} When similarto is specified, use this to remove trace properties to those already existing in the similarto file.

dataproperties_add::Array{DataProperty} When similarto is specfied, use this to add dataset properties to those aloready existing in the similarto file.

dataproperties_rm::Array{DataProperty} When similarto is specified, use this to remove dataset properties to those already existing in the similarto file.

jsopen(filename)

Equivalent to jsopen(filename, "r")

TeaSeis.jsopen_write

TeaSeis.label

TeaSeis.labels — Function.

labels(io)

Returns the string labels corresponding to the framework axes of the JavaSeis dataset corresponding to io::JSeis.

labels(io, i)

Returns the string label of the ith framework axis of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.leftjustify! — Function.

leftjustify(io, trcs, hdrs)

Left justify all live (non-dead) traces in a frame, moving them to the beginning of trcs and hdrs. See also regularize!

TeaSeis.lincs — Function.

lincs(io)

Returns the logical increments of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

lincs(io,i)

Returns the logical increment of the framework axes for dimension i of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.lrange — Function.

lrange(io)

Returns the logical ranges of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

lrange(io, i)

Returns the logical range of the framework axes for dimension i of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.lstarts — Function.

lstarts(io)

Returns the logical start of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

lstarts(io,i)

Returns the logical start of the framework axes for dimension i of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.make_extentdirs

TeaSeis.make_extents

TeaSeis.make_primarydir

TeaSeis.mapposition

TeaSeis.missing

TeaSeis.nextents_heuristic

TeaSeis.packframe!

TeaSeis.packtrace!

TeaSeis.parseindex

TeaSeis.parserng

TeaSeis.parserngs

TeaSeis.pincs — Function.

pincs(io)

Returns the physical increments of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

pincs(io, i)

Returns the physical increments of the framework axes for dimension i of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.prop — Function.

prop(io, propertyname[, proptype=Any])

Get a trace property from io::JSeis where propertyname is either String or TracePropertyDef. Note that if propertyname is a String, then this method produces a type-unstable result. For example:

io = jsopen("data.js")
p = prop(io, "REC_X")            # using a `String`, output type of prop is not inferred
p = prop(io, "REC_X", Float32)   # using a `String`, output type of prop is inferred using `Float32`
p = prop(io, stockprop[:REC_X])  # using a `TracePropertyDef`, output type of prop is inferred

Note that in the examples above, the string "RECX" can be replaced by the symbol `RECX`.

TeaSeis.propdef

TeaSeis.propdefs — Function.

propdefs(io)

Returns the property definitions of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

propdefs(io, i)

Returns the property definition of the ith framework axis of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.propertyformatstring

TeaSeis.proplabel

TeaSeis.props — Function.

props(io)

Returns the trace properties of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

props(io, i)

Returns the trace property of the ith framework axis of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.pstarts — Function.

pstarts(io)

Returns the physical start of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

pstarts(io, i)

Returns the physical start of the ith dimension of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.read_impl!

TeaSeis.readframe — Function.

readframe(io, idx...)

Out-of-place read of a single frame from a JavaSeis dataset. For non full frame, the resulting traces and headers are left justified. Examples:

3D:

frm_idx = 1
trcs, hdrs = readframe(jsopen("data_3D.js"), frm_idx)

4D:

frm_idx, vol_idx = 1, 1
trcs, hdrs = readframe(jsopen("data_4D.js"), frm_idx, vol_idx)

5D:

frm_idx, vol_idx, hyp_idx = 1, 1, 1
trcs, hdrs = readframe(jsopen("data_5D.js"), frm_idx, vol_idx, hyp_idx)

TeaSeis.readframe! — Function.

readframe!(io, trcs, hdrs, idx...)

In-place read of a single frame from a JavaSeis dataset. For non full frame, the resulting traces and headers are left justified. Examples:

3D:

io = jsopen("data_3D.js")
trcs, hdrs = allocframe(io)
frm_idx = 1
readframe!(io, trcs, hdrs, frm_idx)

4D:

io = jsopen("data_4D.js")
trcs, hdrs = allocframe(io)
frm_idx, vol_idx = 1, 1
readframe!(io, trcs, hdrs, frm_idx, vol_idx)

5D:

io = jsopen("data_5D.js")
trcs, hdrs = allocframe(io)
frm_idx, vol_idx, hyp_idx = 1, 1, 1
readframe!(io, trcs, hdrs, frm_idx, vol_idx, hyp_idx)

TeaSeis.readframe_impl

TeaSeis.readframe_impl!

TeaSeis.readframehdrs — Function.

readframehdrs(io, idx...)

Out-of-place read of a single frame (headers only) from a JavaSeis dataset. For non full frame, the resulting headers are left justified. Examples:

3D:

frm_idx = 1
hdrs = readframehdrs(jsopen("data_3D.js"), frm_idx)

4D:

frm_idx, vol_idx = 1, 1
hdrs = readframehdrs(jsopen("data_4D.js"), frm_idx, vol_idx)

5D:

frm_idx, vol_idx, hyp_idx = 1, 1, 1
hdrs = readframehdrs(jsopen("data_5D.js"), frm_idx, vol_idx, hyp_idx)

TeaSeis.readframehdrs! — Function.

readframehdrs!(io, hdrs, idx...)

In-place read of a single frame from a JavaSeis dataset (headers only). For non full frame, the resulting headers are left justified. Examples:

3D:

io = jsopen("data_3D.js")
hdrs = allocframehdrs(io)
frm_idx = 1
readframehdrs!(io, hdrs, frm_idx)

4D:

io = jsopen("data_4D.js")
hdrs = allocframehdrs(io)
frm_idx, vol_idx = 1, 1
readframehdrs!(io, hdrs, frm_idx, vol_idx)

5D:

io = jsopen("data_5D.js")
hdrs = allocframehdrs(io)
frm_idx, vol_idx, hyp_idx = 1, 1, 1
readframehdrs!(io, hdrs, frm_idx, vol_idx, hyp_idx)

TeaSeis.readframehdrs_impl

TeaSeis.readframehdrs_impl!

TeaSeis.readframetrcs — Function.

readframetrcs(io, idx...)

Out-of-place read of a single frame (traces only) from a JavaSeis dataset. For non full frame, the resulting traces are left justified. Examples:

3D:

frm_idx = 1
trcs = readframetrcs(jsopen("data_3D.js"), frm_idx)

4D:

frm_idx, vol_idx = 1, 1
trcs = readframetrcs(jsopen("data_4D.js"), frm_idx, vol_idx)

5D:

frm_idx, vol_idx, hyp_idx = 1, 1, 1
trcs = readframetrcs(jsopen("data_5D.js"), frm_idx, vol_idx, hyp_idx)

TeaSeis.readframetrcs! — Function.

readframetrcs!(io, trcs, hdrs, idx...)

In-place read of a single frame from a JavaSeis dataset (traces only). For non full frame, the resulting traces are left justified. Examples:

3D:

io = jsopen("data_3D.js")
trcs = allocframetrcs(io)
frm_idx = 1
readframetrcs!(io, trcs, frm_idx)

4D:

io = jsopen("data_4D.js")
trcs = allocframetrcs(io)
frm_idx, vol_idx = 1, 1
readframetrcs!(io, trcs, frm_idx, vol_idx)

5D:

io = jsopen("data_5D.js")
trcs = allocframetrcs(io)
frm_idx, vol_idx, hyp_idx = 1, 1, 1
readframetrcs!(io, trcs, frm_idx, vol_idx, hyp_idx)

TeaSeis.readframetrcs_impl

TeaSeis.readframetrcs_impl!

TeaSeis.readhdrs — Function.

readhdrs(io, trace_range, range...)

Out-of-place read of a subset of data (headers only) from a JavaSeis file. Returns an array of trace data. If performance is important, then consider using readframetrcs instead. Examples:

3D:

hdrs = readhdrs(jsopen("data_3D.js"), :, :, :)
hdrs = readhdrs(jsopen("data_3D.js"), :, 1:2:end, 1:5)

4D:

hdrs = readhdrs(jsopen("data_4D.js"), :, :, :, :)
hdrs = readhdrs(jsopen("data_4D.js"), :, :, 2, 2:2:10)

5D:

hdrs = readhdrs(jsopen("data_5D.js"), :, :, :, :, :)
hdrs = readhdrs(jsopen("data_5D.js"), :, :, 2, 2:2:10, 1:10)

TeaSeis.readhdrs! — Function.

readhdrs!(io, hdrs, smp_range, trace_range, range...)

In-place read of a subset of data (headers only) from a JavaSeis file. If performance is important, then consider using readframehdrs! instead. Examples:

3D:

readhdrs!(jsopen("data_3D.js"), hdrs, :, :, :)
readhdrs!(jsopen("data_3D.js"), hdrs, :, 1:2:end, 1:5)

4D:

readhdrs!(jsopen("data_4D.js"), hdrs, :, :, :, :)
readhdrs!(jsopen("data_4D.js"), hdrs, :, :, 2, 2:2:10)

5D:

readhdrs!(jsopen("data_5D.js"), hdrs, :, :, :, :, :)
readhdrs!(jsopen("data_5D.js"), hdrs, :, :, 2, 2:2:10, 1:10)

TeaSeis.readhdrs_impl!

TeaSeis.readmap

TeaSeis.readtrcs — Function.

readtrcs(io, sample_range, trace_range, range...)

Out-of-place read of a subset of data (traces only) from a JavaSeis file. Returns an array of trace data. If performance is important, then consider using readframetrcs instead. Examples:

3D:

trcs = readtrcs(jsopen("data_3D.js"), :, :, :)
trcs = readtrcs(jsopen("data_3D.js"), :, 1:2:end, 1:5)

4D:

trcs = readtrcs(jsopen("data_4D.js"), :, :, :, :)
trcs = readtrcs(jsopen("data_4D.js"), :, :, 2, 2:2:10)

5D:

trcs = readtrcs(jsopen("data_5D.js"), :, :, :, :, :)
trcs = readtrcs(jsopen("data_5D.js"), :, :, 2, 2:2:10, 1:10)

TeaSeis.readtrcs! — Function.

readtrcs!(io, trcs, sample_range, trace_range, range...)

In-place read of a subset of data (traces only) from a JavaSeis file. If performance is important, then consider using readframetrcs! instead. Examples:

3D:

readtrcs!(jsopen("data_3D.js"), trcs, :, :, :)
readtrcs!(jsopen("data_3D.js"), trcs, :, 1:2:end, 1:5)

4D:

readtrcs!(jsopen("data_4D.js"), trcs, :, :, :, :)
readtrcs!(jsopen("data_4D.js"), trcs, :, :, 2, 2:2:10)

5D:

readtrcs!(jsopen("data_5D.js"), trcs, :, :, :, :, :)
readtrcs!(jsopen("data_5D.js"), trcs, :, :, 2, 2:2:10, 1:10)

TeaSeis.readtrcs_impl!

TeaSeis.regularize! — Function.

regularize!(io, trcs, hdrs)

Regularize the traces in a frame, moving them from their left-justified state, to one that reflects their trace location within a frame according to their trace framework definition.

TeaSeis.rmsecondaries

TeaSeis.set! — Function.

set!(prop, hdrs, i, value)

Set the value of the trace property prop::TraceProperty stored in the header of the ith column of hdrs::Array{UInt8,2} to value::T. For example, io=jsopen("test.js"); hdrs=readframehdrs(io,1); set!(prop(io,"REC_X"), 1, 10.0).

TeaSeis.sspropset!

TeaSeis.stockdatatype

TeaSeis.stockdomain

TeaSeis.stockprop

TeaSeis.stockunit

TeaSeis.stringtype2type

TeaSeis.sub2ind

TeaSeis.teaseis_robust_rm

TeaSeis.tracelength

TeaSeis.tracetype

TeaSeis.units — Function.

units(io)

Returns the unit of measure of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

units(io, i)

Returns the unit of measure of the ith dimension of the framework axes of the JavaSeis dataset corresponding to io::JSeis.

TeaSeis.unpackframe!

TeaSeis.unpacktrace!

TeaSeis.volumeindex

TeaSeis.write_extentmanager

TeaSeis.write_fileproperties

TeaSeis.write_helper

TeaSeis.write_nameproperties

TeaSeis.write_parproperty

TeaSeis.write_statusproperties

TeaSeis.write_traceproperty

TeaSeis.write_trcshdrs_helper

TeaSeis.write_virtualfolders

TeaSeis.writeframe — Function.

writeframe(io, trcs, hdrs)

Write a frame of data to the JavaSeis dataset corresponding to io::JSeis. trcs and hdrs are 2-dimensional arrays. The location of the dataset written to is determined by the values of the framework headers stored in hdrs.

writeframe(io, trcs, idx...)

Write a frame of data to the JavaSeis dataset corresponding to io::JSeis. trcs is a 2-dimensional array. The location of the datset written to is determined by idx.... For example:

3D:

writeframe(jsopen("data_3D.js"), trcs, 1) # write to frame 1

4D:

writeframe(jsopen("data_4D.js"), trcs, 1, 2) # write to frame 1, volume 2

5D:

writeframe(jsopen("data_5D.js"), trcs, 1, 2, 3) # write to frame 1, volume 2, hypercube 3

TeaSeis.writeframe_impl