Docstrings

SimplePartitions.Ferrers — Function.

Ferrers(P::IntegerParition) prints a graphical representation of the partition P in the form of a Ferrer's diagram.

SimplePartitions.IntegerPartition — Type.

An IntegerPartition is a decreasing list of positive integers. Construct in the following ways:

IntegerParition([a,b,c,...])

IntegerPartition(a,b,c,...)

SimplePartitions.Partition — Type.

A Partition is a set of nonempty, pairwise disjoint sets. A new Partition is created by specifying the ground set A and calling Partition(A). The set A may be either a Set{T} for some type T or an BitSet.

The parameter A may also be a list (one-dimensional array).

In addition, Partition(n) for a nonnegative integer n creates a partition of the set {1,2,...,n}.

The datatype Partition is, essentially, a wrapper around the DataStructures.DisjointSets type.

SimplePartitions.PartitionBuilder — Function.

PartitionBuilder(A,check=true) takes a set of nonempty, pairwise disjoint sets and creates the corresponding partition. It is the inverse operation to parts(P). The optional parameter check causes sanity checks to be return on the input set of sets (throwing errors if it is invalid).

SimplePartitions.SimplePartitions

SimplePartitions.all_partitions — Function.

all_partitions(A::Set) creates a Set containing all possible partitions of the set A.

all_partitions(n::Int) creates the Set of all partitions of the set {1,2,...,n}.

Both of these take an optional second argument k to specify that only partitions with exactly k parts should be returned.

SimplePartitions.build_sos!

SimplePartitions.eval

SimplePartitions.find_part — Function.

find_part(P,a) returns the part of P that contains a (or throws an error if a is not in the ground set).

SimplePartitions.ground_set — Function.

elements(P) returns (a copy of) the ground set of the partition P.

SimplePartitions.in_same_part — Function.

in_same_part(P,a,b) returns true if a and b are in the same part of the partition P. An error is thrown if either is not in the ground set of P.

SimplePartitions.include

SimplePartitions.kill_sos! — Function.

kill_sos!(P) wipes out the set-of-sets cache. Only used internally.

SimplePartitions.meet — Function.

meet(P,Q) computes the meet of the partitions. This may also be invoked as P*Q.

SimplePartitions.merge_parts! — Function.

merge_parts!(P,a,b) updates P by merging the parts that contain elements a and b.

merge_parts!(P,elts) merges all the elements in elts into a part. Thus merge_parts!(P,[a,b]) is equivalent to merge_parts!(P,a,b).

SimplePartitions.num_elements — Function.

num_elements(P) gives the number of elements in the ground set of the partition P. This equals the sum of the sizes of the parts.

SimplePartitions.num_parts — Function.

num_parts(P) gives the number of parts in the partition P.

SimplePartitions.parts — Function.

parts(P) returns a set containing the parts of the partition P. That is, we return a set of sets.

SimplePartitions.refines — Function.

refines(P,Q) determines if P is a refinement of Q. That is, is every part of P a subset of a part of Q? The two partitions must have the same ground set of else an error is thrown.

refines(P,Q) can be invoked as P<=Q. The variants P<Q, P>=Q, and P>Q operate as expected. Note that partitions are only partially ordered by refinement and one can easily construct partitions P and Q for which both P<=Q and Q<=P are false.

SimplePartitions.set_element_type — Function.

set_element_type(A) gives the element type of a set A.

SimplePartitions.∧

SimplePartitions.∨