Docstrings

CancerSeqSim.AllMetrics

CancerSeqSim.AnalysedData

CancerSeqSim.CancerCells

CancerSeqSim.CancerNonStemCell

CancerSeqSim.CancerSeqSim

CancerSeqSim.CancerStemCell

CancerSeqSim.Draw

CancerSeqSim.InputParameters

CancerSeqSim.InputParametersM

CancerSeqSim.Mcdf

CancerSeqSim.MetricObj

CancerSeqSim.RawOutput

CancerSeqSim.RawResults

CancerSeqSim.RsqObj

CancerSeqSim.SampledData

CancerSeqSim.SimResult

CancerSeqSim.SimResultM

CancerSeqSim.Simulation

CancerSeqSim.SimulationM

CancerSeqSim.SimulationSamples

CancerSeqSim.SimulationStemCells

CancerSeqSim.Stem

CancerSeqSim.StemCellSimResult

CancerSeqSim.allelefreq

CancerSeqSim.allelefreqexpand

CancerSeqSim.areametric

CancerSeqSim.areametricraw

CancerSeqSim.bdprocess

CancerSeqSim.betabinom

CancerSeqSim.birthdeathprocess

CancerSeqSim.calculateclonefreq

CancerSeqSim.cancercell

CancerSeqSim.cancercellM

CancerSeqSim.cellsconvert

CancerSeqSim.copycell

CancerSeqSim.cumulativedist

CancerSeqSim.cumulativeplot

CancerSeqSim.eval

CancerSeqSim.exptime

CancerSeqSim.getmetrics

CancerSeqSim.getresults

CancerSeqSim.getsamples

CancerSeqSim.getsummary

CancerSeqSim.include

CancerSeqSim.initializesim

CancerSeqSim.initializesimsample

CancerSeqSim.kolmogorovD

CancerSeqSim.kolmogorovDmean

CancerSeqSim.meanDraw

CancerSeqSim.meantime

CancerSeqSim.metricp

CancerSeqSim.multiplicativefitness

CancerSeqSim.newmutations

CancerSeqSim.newmutations2

CancerSeqSim.newmutationsinit

CancerSeqSim.newmutationsinitsample

CancerSeqSim.newmutationssample

CancerSeqSim.newnonstemcell

CancerSeqSim.nonmultiplicativefitness

CancerSeqSim.nonstemcelldivision

CancerSeqSim.rsq

CancerSeqSim.run1simulation

CancerSeqSim.run1simulationsample

CancerSeqSim.run1simulationstem

CancerSeqSim.sampledhist

CancerSeqSim.sampledhistmultiple

CancerSeqSim.selection

CancerSeqSim.selection2clone

CancerSeqSim.sfunc

CancerSeqSim.simulate — Function.

simulate(; <keyword arguments>)

Simulate a stochastic model of tumour growth with a single subclone introduced at a random time and with a random fitness advantage. Output return synthetic sequencing data. ...

Arguments

read_depth = 200.0: Mean read depth of the target data set

detectionlimit = 5/read_depth: Ability to detect low frequency variants. Assumes 5 reads are needed to call a variant.

μ = 10.0: Mutation rate per division per genome (this will timesed by ploidy for the mutation rate per cell). At each division a Poisson random variable with mean μ is sampled.

clonalmutations = 100.0: Number of clonal mutations present in the first cell.

nclones = 1: Number of subclones introduced

Nmax = 10^4: Maximum population size.

ρ = 0.0: Overdispersion parameter for beta-binomial model of sequencing data. ρ = 0.0 means model is binomial sampling

timefunction = exptime: Function for KMC algorithm timestep. exptime returns an exponentially distributed random variable, if you would rather return the mean of the distribution define a function that returns 1. ie returnone() = 1.

ploidy = 2: ploidy of the genome

d = 0.0: Death rate of the thost population in the tumour

b = log(2): Birth rate of the population. Set to log(2) so that tumour doubles with each unit increase in t in the absence of cell death.

fixedmu = false: If set to false number of mutations per division is fixed and not sampled from a poisson distribution.

...

simulate(minclonesize, maxclonesize; <keyword arguments>)

Return simulation with frequency of subclones >minclones & <maxclonesize.

simulate(minclonesize, maxclonesize, independentclones::Bool; <keyword arguments>)

Return simulation with frequency of subclones >minclones & <maxclonesize and specify whether subclones are independent or not (ie nested or not). Only applicable to >1 subclone.

simulate(minclonesize, maxclonesize, mindiff; <keyword arguments>)

Return simulation with frequency of subclones >minclones & <maxclonesize and subclone frequency are at least mindiff apart.

simulate(minclonesize, maxclonesize, mindiff; <keyword arguments>)

Return simulation with frequency of subclones >minclones & <maxclonesize and subclone frequency are at least mindiff and have at least minmutations mutations.

CancerSeqSim.simulatedifferentmutations

CancerSeqSim.simulatedifferentmutationsmoran

CancerSeqSim.simulatestemcells — Function.

simulatestemcells(; <keyword arguments>)

Simulate a stochastic model of tumour growth with a stem cell architecture. Output return synthetic sequencing data. ...

Arguments

α = 0.1: Symmetric division rate

maxdivisions = 5: Maximum number of divisions of differentiated cells.

read_depth = 200.0: Mean read depth of the target data set

detectionlimit = 5/read_depth: Ability to detect low frequency variants. Assumes 5 reads are needed to call a variant.

μ = 10.0: Mutation rate per division. At each division a Poisson random variable with mean μ is sampled.

clonalmutations = 100.0: Number of clonal mutations present in the first cell.

Nmax = 10^4: Maximum population size.

ρ = 0.0: Overdispersion parameter for beta-binomial model of sequencing data. ρ = 0.0 means model is binomial sampling

timefunction = exptime: Function for KMC algorithm timestep. exptime returns an exponentially distributed random variable, if you would rather return the mean of the distribution define a function that returns 1. ie returnone() = 1.

ploidy = 2: ploidy of the genome

d = 0.0: Death rate of the thost population in the tumour

b = log(2): Birth rate of the population. Set to log(2) so that tumour doubles with each unit increase in t in the absence of cell death.

fixedmu = false: If set to false number of mutations per division is fixed and not sampled from a poisson distribution.

...

CancerSeqSim.stemcelldivision

CancerSeqSim.trapz

CancerSeqSim.tumourgrow_birthdeath

CancerSeqSim.tumourgrow_birthdeathsample

CancerSeqSim.tumourgrow_stemcell

CancerSeqSim.tumourmoran

CancerSeqSim.vafhistogram

CancerSeqSim.x