Docstrings

Plotly.@L_mstr

Plotly.@L_str

Plotly.API_ROOT

Plotly.AbstractLayout

Plotly.AbstractTrace

Plotly.ApiCall

Plotly.Cycler

Plotly.DEFAULT_CONFIG

Plotly.GenericTrace

Plotly.Layout

Plotly.Plot

Plotly.Plotly

Plotly.PlotlyAPIError

Plotly.PlotlyBase

Plotly.PlotlyConfig

Plotly.PlotlyCredentials

Plotly.PlotlyError

Plotly.PlotlyJS

Plotly.RemotePlot — Type.

Proxy for a plot stored on the Plotly cloud.

Plotly.STYLES

Plotly.Shape

Plotly.Style

Plotly._SRC_ATTRS

Plotly._VERSION

Plotly._get_uid

Plotly.add_recession_bands!

Plotly.addtraces

Plotly.addtraces!

Plotly.api_url

Plotly.area

Plotly.attr

Plotly.bar

Plotly.basic_auth

Plotly.box

Plotly.candlestick

Plotly.carpet

Plotly.choropleth

Plotly.circle

Plotly.comment_create

Plotly.comment_create_raw

Plotly.comment_delete

Plotly.comment_delete_raw

Plotly.cone

Plotly.contour

Plotly.contourcarpet

Plotly.dashboard_create

Plotly.dashboard_create_raw

Plotly.dashboard_list

Plotly.dashboard_list_raw

Plotly.dashboard_partial_update

Plotly.dashboard_partial_update_raw

Plotly.dashboard_permanent_delete

Plotly.dashboard_retrieve

Plotly.dashboard_retrieve_raw

Plotly.dashboard_schema

Plotly.dashboard_schema_raw

Plotly.dashboard_trash

Plotly.dashboard_trash_raw

Plotly.dashboard_update

Plotly.dashboard_update_raw

Plotly.dashboard_writeable_metadata

Plotly.deletetraces

Plotly.deletetraces!

Plotly.download_image

Plotly.download_plot

Plotly.eval

Plotly.extendtraces

Plotly.extendtraces!

Plotly.extra_content

Plotly.extra_content_raw

Plotly.extra_create

Plotly.extra_create_raw

Plotly.extra_delete

Plotly.extra_detail

Plotly.extra_detail_raw

Plotly.extra_partial_update

Plotly.extra_partial_update_raw

Plotly.extract_grid_data

Plotly.extract_grid_data! — Function.

srcify!(p::Plot; fileopt::Symbol=:overwrite, grid_fn=nothing, kwargs...)

Look through each trace and the Layout for attributes that have a value of type Union{AbstractArray,Tuple} and are able to be set via (attributename)(src) in the plotly.js api. For each of these, do the following:

Extract the value so it can become a column in a Grid

Remove that key/value pair from the trace/Layout

This happens in place, so the srcified fields will be removed within the input to this function.

Plotly.fid — Function.

fid(rp::RemotePlot)

Return the unique plotly fid for rp. Throws an error if the url inside rp is not correctly formed.

Plotly.file_content

Plotly.file_content_raw

Plotly.file_copy

Plotly.file_copy_raw

Plotly.file_drop_reference

Plotly.file_drop_reference_raw

Plotly.file_image

Plotly.file_image_raw

Plotly.file_lookup

Plotly.file_lookup_raw

Plotly.file_partial_update

Plotly.file_partial_update_raw

Plotly.file_path

Plotly.file_path_raw

Plotly.file_permanent_delete

Plotly.file_remove_star

Plotly.file_restore

Plotly.file_restore_raw

Plotly.file_retrieve

Plotly.file_retrieve_raw

Plotly.file_sources

Plotly.file_sources_raw

Plotly.file_star

Plotly.file_star_raw

Plotly.file_trash

Plotly.file_trash_raw

Plotly.file_update

Plotly.file_update_raw

Plotly.file_writeable_metadata

Plotly.folder_all

Plotly.folder_all_raw

Plotly.folder_create

Plotly.folder_create_raw

Plotly.folder_detail

Plotly.folder_detail_raw

Plotly.folder_home

Plotly.folder_home_raw

Plotly.folder_permanent_delete

Plotly.folder_permanent_delete_raw

Plotly.folder_restore

Plotly.folder_restore_raw

Plotly.folder_shared

Plotly.folder_shared_raw

Plotly.folder_starred

Plotly.folder_starred_raw

Plotly.folder_trash

Plotly.folder_trash_raw

Plotly.folder_trashed

Plotly.folder_trashed_raw

Plotly.fork

Plotly.get_config — Function.

get_config()

Return the session configuration if defined –> otherwise use .config specs

Plotly.get_config_file — Function.

get_config_file()

Load endpoint configuration as a Dict

Plotly.get_credentials — Function.

get_credentials()

Return the session credentials if defined –> otherwise use .credentials specs

Plotly.get_credentials_env

Plotly.get_credentials_file — Function.

get_credentials_file()

Load user credentials informaiton as a dict

Plotly.get_headers

Plotly.get_json

Plotly.get_json_data

Plotly.get_req_headers

Plotly.grid_content

Plotly.grid_content_raw

Plotly.grid_create

Plotly.grid_create_raw

Plotly.grid_destroy

Plotly.grid_drop_reference

Plotly.grid_drop_reference_raw

Plotly.grid_get_col

Plotly.grid_get_col_raw

Plotly.grid_lookup

Plotly.grid_lookup_raw

Plotly.grid_overwrite! — Function.

grid_overwrite(cols::AbstractDict; fid::String="", path::String="")

Replace the data in the grid assocaited with fid fid or at the path path with data in cols. cols should be an AbstractDict mapping from column names to column data. The output of this function is an AbstractDict mapping from column names to column uids in the updated grid.

There are three possible scenarios for the data:

The column appears both in the grid and in cols. In this case the data in that column of the grid will be updated to match the data in cols

The column appears only in cols. In this case a new column will be created in the grid

The column appears only in the grid. Nothing happens...

NOTE: only one of fid and path can be passed

Plotly.grid_partial_update

Plotly.grid_partial_update_raw

Plotly.grid_permanent_delete

Plotly.grid_post_col

Plotly.grid_post_col_raw

Plotly.grid_put_col

Plotly.grid_put_col_raw

Plotly.grid_restore

Plotly.grid_restore_raw

Plotly.grid_retrieve

Plotly.grid_retrieve_raw

Plotly.grid_row

Plotly.grid_trash

Plotly.grid_trash_raw

Plotly.grid_update

Plotly.grid_update_raw

Plotly.grid_writable_metadata

Plotly.heatmap

Plotly.heatmapgl

Plotly.histogram

Plotly.histogram2d

Plotly.histogram2dcontour

Plotly.hline

Plotly.html_body

Plotly.image_generate

Plotly.image_generate_raw

Plotly.include

Plotly.json

Plotly.line

Plotly.make_method

Plotly.mesh3d

Plotly.movetraces

Plotly.movetraces!

Plotly.ohlc

Plotly.openurl

Plotly.parcoords

Plotly.path

Plotly.pie

Plotly.plot

Plotly.plot_content

Plotly.plot_content_raw

Plotly.plot_create

Plotly.plot_create_raw

Plotly.plot_detail

Plotly.plot_detail_raw

Plotly.plot_feed

Plotly.plot_feed_raw

Plotly.plot_list

Plotly.plot_list_raw

Plotly.plot_partial_update

Plotly.plot_partial_update_raw

Plotly.plot_schema_get

Plotly.plot_schema_get_raw

Plotly.plot_update

Plotly.plot_update_raw

Plotly.plotlyconfig

Plotly.plotlycredentials

Plotly.pointcloud

Plotly.post

Plotly.post_v2 — Function.

Post a local Plotly plot to the Plotly cloud using V2 api.

Must be signed in first. See Plotly.signin for details on how to do that

Plotly.prependtraces

Plotly.prependtraces!

Plotly.purge!

Plotly.react

Plotly.react!

Plotly.rect

Plotly.redraw

Plotly.redraw!

Plotly.relayout

Plotly.relayout!

Plotly.request

Plotly.request_data

Plotly.restyle

Plotly.restyle!

Plotly.sankey

Plotly.savefig

Plotly.savefig_remote

Plotly.savejson

Plotly.scatter

Plotly.scatter3d

Plotly.scattercarpet

Plotly.scattergeo

Plotly.scattergl

Plotly.scattermapbox

Plotly.scatterpolar

Plotly.scatterpolargl

Plotly.scatterternary

Plotly.search_list

Plotly.search_list_raw

Plotly.set_config_file — Function.

set_config_file(input_config::AbstractDict)

Save Plotly endpoint configuration as JSON key-value pairs in userhome/.plotly/.config. This includes the plotlydomain, and plotlyapi_domain.

set_config_file(config::PlotlyConfig)

Set the values in the configuration file to match the values in config

Plotly.set_credentials_file — Function.

set_credentials_file(input_creds::AbstractDict)

Save Plotly endpoint configuration as JSON key-value pairs in userhome/.plotly/.credentials. This includes username and api_key.

Plotly.signin — Function.

signin(username::String, api_key::String, endpoints=nothing)

Define session credentials/endpoint configuration, where endpoint is a Dict

Plotly.splom

Plotly.srcify — Function.

srcify(p::Plot)

Allocating version of srcify! the plot will be deepcopied before passing to srcify!, so the argument passed to this function will not be modified.

Plotly.srcify! — Function.

srcify!(p::Plot; fileopt::Symbol=get_config().fileopt, grid_fn=nothing, kwargs...)

This function does three things:

Calls extract_grid_data!(p) (see docs) to remove attributes that can be set via (attributename)(src) in the plotly.js api.

Creates a grid on the plotly server containing the extract data

Maps the (attributename)(src) attribute to the grid column

If fileopt is :create and grid_fn exists under the User's plotly account, then the changes described above will happen in-place on the grid.

If either of those conditions are not met, then a new grid will be created.

Plotly.stem

Plotly.streamtube

Plotly.style

Plotly.surface

Plotly.table

Plotly.to_image

Plotly.try_lookup

Plotly.try_me

Plotly.update

Plotly.update!

Plotly.use_style!

Plotly.validate_response

Plotly.violin

Plotly.vline