Docstrings

ProteinEnsembles.Atom — Type.

An atom from a PDB format file.

ProteinEnsembles.BondedPair — Type.

A pair of atoms with a covalent bond between them in the standard amino acids.

ProteinEnsembles.Constraints — Type.

Information about the lower and upper distance constraints between atoms.

ProteinEnsembles.ModelledEnsemble — Type.

A group of modelled structures.

ProteinEnsembles.ModelledStructure — Type.

A set of coordinates and a score.

ProteinEnsembles.PCA — Type.

The results of a principal components analysis, sorted with highest eigenvalue first.

ProteinEnsembles.ProteinEnsembles

ProteinEnsembles.addstructure! — Function.

Add a structure to an ensemble.

ProteinEnsembles.align! — Function.

Align the first set of atomic coordinates to the second using an iterative procedure. At each step atoms with a high structural deviation are removed. If the standard deviation of the structural deviations is greater than cutoff, then structural deviations further than 2 standard deviations from the median are discarded. This is repeated n_cycles times. If calpha is true, only considers the C-alpha atoms.

ProteinEnsembles.alignatoms! — Function.

Align an array of atoms to a set of coordinates or another set of atoms. Input is list of atoms and coords to align to. Changes the atom coords.

ProteinEnsembles.alignensemble! — Function.

Align an ensemble of coordinates to a coordinate set using an iterative procedure for each structure.

ProteinEnsembles.alignensemblesimple! — Function.

Align an ensemble of coordinates to a coordinate set using a single Kabsch rotation for each structure.

ProteinEnsembles.alignsimple! — Function.

Align the first set of atomic coordinates to the second using a single Kabsch rotation. If calpha is true, only considers the C-alpha atoms.

ProteinEnsembles.angles — Function.

Calculate the bond angle matrix for a list of atoms. Returns a 2D boolean array of whether bond angles are present.

ProteinEnsembles.areatomssame — Function.

Determines if two atoms are the same. Atoms are the same if they have the same atom_name, chain_id, res_n and element. res_name is not considered.

ProteinEnsembles.atomcoords — Function.

Get or set the list of coordinates for a list of atoms.

ProteinEnsembles.atomcoords! — Function.

Set the coordinates of a list of atoms from a list of coordinates.

ProteinEnsembles.atomid — Function.

A string describing an Atom in the form atom_name/res_name/res_n/chain_id.

ProteinEnsembles.atommap — Function.

Finds the mapping of the atoms in the first list to those in the second. Returns an array of length the same as the second list where the number is the atom index in the first list. Atoms missing from the first list are given 0.

ProteinEnsembles.bonds — Function.

Calculate the bonding matrix for a list of atoms. Returns a 2D boolean array of bonding.

ProteinEnsembles.calphaindices — Function.

Return as an array the indices in an array of atoms that correspond to C-alpha atoms.

ProteinEnsembles.centroid — Function.

Find the average structure in an ensemble. The ensemble should be aligned.

ProteinEnsembles.checkfilepath — Function.

Checks if an output filepath is valid, and throws an error if not.

ProteinEnsembles.clusterligsite — Function.

Assign pocket numbers to LIGSITEcs pocketr.pdb file points. Arguments are pocketr.pdb file, pocketall.pdb file and new pocketr.pdb file. New file has pocket number in place of residue number. Apears to agree largely but not completely with the LIGSITEcs clustering. Unassigned points given number 0 and should be fixed manually.

ProteinEnsembles.default_colours — Constant.

Dictionary of default colours for plotting.

ProteinEnsembles.defaults — Constant.

Dictionary of default values for parameters.

ProteinEnsembles.displacements — Function.

Return the distances between each point for two sets of aligned coordinates.

ProteinEnsembles.distances — Function.

Calculate the distance matrix for a list of atoms. Returns a 2D array of distances.

ProteinEnsembles.divalents — Function.

Calculate the divalent matrix for a list of atoms. Returns a 2D boolean array of whether divalents are present.

ProteinEnsembles.ensemblermsds — Function.

Get the RMSD between each structure in an ensemble and a reference structure.

ProteinEnsembles.errorscore — Function.

Calculate SPE error score.

ProteinEnsembles.eval

ProteinEnsembles.fillconstraints — Function.

Use stochastic proximity embedding to generate a structure iteratively. See Agrafiotis et.al., Chapter 14 in Mucherino et. al., 2013 for algorithm details (Algorithm 8). Returns coordinate list.

ProteinEnsembles.findcommonatoms — Function.

Checks two atom arrays and finds common atoms. Returns copies of input atom arrays with only common atoms remaining. The returned arrays are therefore of equal length.

ProteinEnsembles.findimportantpcs — Function.

Takes two sets of projections onto the PCs of a PCA and returns the PC indices ordered such that the first represents the largest difference between the two structures.

ProteinEnsembles.findinteractions — Function.

Find the interactions between atoms in a list of atoms. Returns 2D array of interactions by integer label. Interactions the same as De Groot et. al., Structure, 1997, but cis/trans limits not implemented.

ProteinEnsembles.fixchirality! — Function.

Checks the chirality of coordinate set and switches it if it isn't correct. The coordinates must correspond to the atoms. Inverts each z coordinate in the coordinate array if required.

ProteinEnsembles.fixstring — Function.

Converts a string, character or number to a fixed length string. Be careful - could e.g. cut 0.5 to "0". Also does not round correctly.

ProteinEnsembles.fluctuations — Function.

Gets mean square fluctuation for each atom in an ensemble. Assumes the ensemble is aligned sensibly. Returns a list of mean square fluctuations.

ProteinEnsembles.fractionbetweeninputs — Function.

Fraction of structures in an ensemble that are closer to both input structures than the input structures are to each other.

ProteinEnsembles.generateensemble — Function.

Generate sets of coordinates with the correct chirality from distance constraints. discardratio gives the fraction of extra structures to calculate, then the low-scoring structures are discarded. If `removemodis true then modulator atoms and coords are removed from the ensemble before it is returned. Returns aModelledEnsemble`.

ProteinEnsembles.hard_sphere_radius — Constant.

Dictionary of atomic hard sphere radii in Angstroms.

ProteinEnsembles.include

ProteinEnsembles.inferelement — Function.

Determines the element of an atom from the atom name. Returns the element as a string.

ProteinEnsembles.interactioninfo — Function.

Returns a list with interaction types and a list with corresponding distance tolerances. Zero index represents same atom. tolerance_weight weights all tolerances.

ProteinEnsembles.interactions — Function.

Run the interaction finding pipeline with one or two structures. If one structure is given, returns a Constraints object. If two structures are given, returns three Constraints objects corresponding to the combined constraints, constraints from structure one and constraints from structure two.

ProteinEnsembles.isdbonefour — Function.

Determines if two atoms are a 1-4 pair around a restricted side chain double bond. Assumes atoms are a 1-4 pair.

ProteinEnsembles.ishbond — Function.

Determines if two atoms are part of a hydrogen bond where the donor-acceptor distance is <= 3.5 Angstroms, the hydrogen-acceptor distance is <= 2.5 Angstroms and the donor-hydrogen-acceptor angle is minimally 90 degrees. Won't pick up the donor-H pairing, even though it is part of the H bond, as this pairing is covalent. Sulfur not considered.

ProteinEnsembles.ishydrophobic — Function.

Determines if two atoms are closer than the sum of the VDW radii plus extradist Angstroms. Currently counting C and H atoms only. Applies a crude 5 A distance cutoff for speed, so extradist values above ~1.5 A will have no effect.

ProteinEnsembles.isloosephipsi — Function.

Determines if two atoms in a 1-4 pair around the phi/psi angle are either in a loop region or one of the residues is glycine. Assumes atoms are a 1-4 pair around the phi/psi angle.

ProteinEnsembles.ismodulator — Function.

Determines whether an Atom is a fake modulator atom.

ProteinEnsembles.isomonefour — Function.

Determines if two atoms are a 1-4 pair around the omega angle. Assumes atoms are a 1-4 pair.

ProteinEnsembles.isphipsi — Function.

Determines if two atoms are a 1-4 pair around the phi/psi angle. Assumes atoms are a 1-4 pair. Glycine H is not polar so should not be present.

ProteinEnsembles.isring — Function.

Determines if two atoms are part of the same ring system.

ProteinEnsembles.issaltbridge — Function.

Determines if two atoms are part of oppositely charged groups that are within 4 Angstroms. His not included.

ProteinEnsembles.issecstr — Function.

Determines if two atoms are backbone atoms that are part of the same helix/strand and not more than 4 residues apart. Does not include 3-10 and pi helices.

ProteinEnsembles.istightphipsi — Function.

Determines if two atoms in a 1-4 pair around the phi/psi angle are either in the same helix/strand or one of the residues is proline. Assumes atoms are a 1-4 pair around the phi/psi angle.

ProteinEnsembles.kabschalignment — Function.

Find the transformations to map the first set of coordinates onto the second. See Kabsch algorithm for details. Returns the translation to center the first set, the translation to center the second set and the rotation to map the first set onto the second.

ProteinEnsembles.makedirectories — Function.

Make the outer results directory and any inner directories from a list.

ProteinEnsembles.mod_atom_info — Constant.

Dictionary of atomic information for fake modulator atoms.

ProteinEnsembles.newcoords

ProteinEnsembles.paramfromshell — Function.

Run auto-parameterisation from command line.

ProteinEnsembles.parampipeline — Function.

Run the auto-parameterisation pipeline.

ProteinEnsembles.parsecommandlineparam — Function.

Read arguments from command line for an ExProSE auto-parameterisation.

ProteinEnsembles.parsecommandlinerun — Function.

Read arguments from command line for an ExProSE run.

ProteinEnsembles.pcsaddlayer! — Function.

Adds a layer of points to an existing layer for PC plotting.

ProteinEnsembles.perturbensemble — Function.

Gets ensembles for multiple modulators. Use pocket points to generate additional constraints.

ProteinEnsembles.plotfluctuations — Function.

Plot root mean square fluctuations of an aligned ensemble. Arguments are the output filepath and the array of fluctuations.

ProteinEnsembles.plotpcs — Function.

Plot the projections of an ensemble onto the principal components of the motion.

ProteinEnsembles.plottheme — Function.

Theme for plots.

ProteinEnsembles.pointtheme — Function.

Theme for plot points.

ProteinEnsembles.projectensemble — Function.

Project the structures in an ensemble onto the eigenvectors of a PCA to get the principal coordinates. Non-zero eigenvalues only are returned. Assumes the ensemble is aligned and corresponds exactly to the PCA.

ProteinEnsembles.projectpdbfile — Function.

Projects the structure in a PDB file onto the eigenvectors of a PCA. The residue numbering between the PDB files must be consistent but missing/extra residues are allowed. Reads the PDB file, maps the atoms to the ref atoms, aligns the atoms to the ref atoms then gets the projections. Arguments are PDB filepath, reference atoms to get atom map, coords to align to and the PCA object.

ProteinEnsembles.projectpdbfiles — Function.

Wrapper function for projectpdbfile to project multiple PDBs and write the output.

ProteinEnsembles.projectstructure — Function.

Project a list of atoms onto the eigenvectors of a PCA. Atom map is the mapping of the input atom indices to the reference atoms. It is not required if the atoms correspond to the PCA. calpha_inds is only required if atom_map is non-trivial and is a list of C-alpha indices. 0 means the atom is not in the reference atoms. Assumes the structure is aligned.

ProteinEnsembles.protein_bonds — Constant.

A list of bonded pairs representing the covalent bonds in a standard protein.

ProteinEnsembles.readdssp — Function.

Reads a DSSP file. If atoms are provided, creates additional '-' entries for residues not present in the DSSP file. Returns a dictionary where key is residue number and chain, and value is secondary structure.

ProteinEnsembles.readensemble — Function.

Read an ensemble of structures and return a ModelledEnsemble.

ProteinEnsembles.readligsite — Function.

Read a LIGSITEcs PDB output file. Returns the pocket centre coordinate list and list of pocket volumes.

ProteinEnsembles.readpdb — Function.

Parse a PDB file. Returns a list of atoms. hetatm determines whether HETATM records are read in.

ProteinEnsembles.readpdblines — Function.

Read the ATOM lines from a PDB file into an array.

ProteinEnsembles.readpocketpoints — Function.

Read a custom pocket point file in PDB format where the residue number is the pocket number. This file can be generated using a script from the LIGSITEcs pocketr.pdb and pocketall.pdb files. Returns a dictionary where key is pocket number and value is point coords.

ProteinEnsembles.removemodulator! — Function.

Takes a ModelledEnsemble and removes the fake modulator atoms and coords.

ProteinEnsembles.repeatpocketpoints — Function.

Repeat pocket point to get a certain number of points. Points are repeated as long as whole copies can be made, then the remainder are chosen randomly. Returns coordinate array.

ProteinEnsembles.rmsd — Function.

Get the RMSD between two identical structures or two sets of coordinates. Assumes the structures are aligned. If calpha is true, only considers the C-alpha atoms.

ProteinEnsembles.runanalysis — Function.

Run the analysis pipeline and write output files.

ProteinEnsembles.runfromshell — Function.

Run pipeline from command line.

ProteinEnsembles.runpipeline — Function.

Wrapper function to run the whole pipeline.

ProteinEnsembles.selfalignensemble! — Function.

Align an ensemble to itself. See Bakan and Bahar, PNAS, 2009 for algorithm. The single Kabsch alignment is used during iteration, then the iterative alignment is done at the end. Returns the structure the final alignment was on, not the ensemble average (but they will be close). The cutoff parameter determines the RMSD between two average structures below which iteration is terminated.

ProteinEnsembles.spaceatomname — Function.

Space an atom name correctly in the four-character limit. The element is in position 2. Returns a string of length 4.

ProteinEnsembles.tmscore — Function.

TM-score of model and reference, given as filepaths. Uses external command line program - see https://zhanglab.ccmb.med.umich.edu/TM-score.

ProteinEnsembles.tmscorepathvalid — Function.

Check a command runs TMscore.

ProteinEnsembles.trimbyscore! — Function.

Retains only the ntokeep top scoring coordinate sets in an ensemble.

ProteinEnsembles.vdw_radius — Constant.

Dictionary of atomic VDW radii in Angstroms.

ProteinEnsembles.writeclusterpoints — Function.

Write output file with assignment in residue number column

ProteinEnsembles.writeensemble — Function.

Write an ensemble out as separate PDB files.

ProteinEnsembles.writeensemblescores — Function.

Write out the scores associated with the structures in an ensemble.

ProteinEnsembles.writefloatarray — Function.

Write an array of floats to an output file.

ProteinEnsembles.writeintarray — Function.

Write an array of integers to an output file.

ProteinEnsembles.writepcview — Function.

Write a PyMol script to view the atom motions of a principal component.

ProteinEnsembles.writepcviews — Function.

Wrapper function for writepcview that produces visualisation scripts for the top n_to_write PCs.

ProteinEnsembles.writepdb — Function.

Write a list of atoms as a PDB file. The atoms are not reordered so it is assumed they are in a sensible order, e.g. the order they were read in. Only ATOM records are written; no TER labels. Could write a brief header for generated models.

ProteinEnsembles.writeprojections — Function.

Write out the projections of structures onto the eigenvectors of a PCA. pcstowrite is a list of the PCs to write, e.g. [1, 2]. All PCs written by default.

ProteinEnsembles.writestringarray — Function.

Write an array of strings to an output file.