Docstrings

Hermetic.He_coefficients

Hermetic.He_value — Function.

He_value(n, x)

evaluates He(n,x) polynomial

Args:

o::Int the degree of the polynomial

x::Array{Float64, 1} the evaluation points

Hermetic.Hen_coefficients — Function.

Hen_coefficients(n)

Calculate the coefficient of Hen(n, x), where Hen(n, x) is the normalized polynomial probabilist Hermite polynomial of degree n

Details:

The normalized probabilist Hermite polynomails is defined as

$\int H{enj}(x)H{enk}(x) w(x) dx = \delta_{jk}$

where w(x) is the normal density function.

Args:

n the order of the polynomial

Output:

n the order of the polynomial

c the coefficients

e the exponents of the polynomial

Hermetic.Hen_value — Function.

Hen_value(n, x)

evaluates Hen(n,x) polynomial

Args:

o::Int the degree of the polynomial

x::Array{Float64, 1} the evaluation points

Hermetic.Henp_to_polynomial — Function.

Henp_to_polynomial (m::Int, l::Array{Int, 1})

writes a Hermite Product Polynomial as a standard polynomial.

Details:

Hen(i,x) represents the probabilist's normalized Hermite polynomial.

For example, if M = 3, L = [1, 0, 2 ]

then

Hen(1,0,2)(X,Y,Z) = Hen(1)(X) * Hen(0)(Y) * Hen(2)(Z) = X * 0.707107*(Y^2-1) *

Args:

m::Int the spatial dimension of the hermite's normalized product (Hnp) polynomial

l::Array{Int, 1} the index of the Hnp

Output:

o::Int the "order" of the polynomial product

c::Array{int,1} the coefficients of the polynomial product

e::Array{Int,1} the indices of the exponents of the polynomial product.

Hermetic.Henp_to_polynomial_fullw

Hermetic.Hermetic

Hermetic.Hermite

Hermetic.PolyType

Hermetic.ProductPoly

Hermetic.Standard

Hermetic._set_ppoly

Hermetic.eval

Hermetic.expectation_monomial — Function.

calculate the expectation of the monomial with exponent e with respect to n(0,1)

Hermetic.expectation_monomial!

Hermetic.fall_fact — Function.

fall_fact(x, n)

Input, int X, the argument of the falling factorial function.

Input, int N, the order of the falling factorial function.
If N = 0, FALL = 1, if N = 1, FALL = X.  Note that if N is
negative, a "rising" factorial will be computed.

Hermetic.gamma_half_integer — Function.

gammahalfinteger(j)

Calculate Gamma(j/2)/√π

Hermetic.get_inter_idx — Function.

getinteridx(X::Array{Int, 2})

Return the index of interaction terms of order equal to Ki of a multivariate polynomial of order k and dimension m.

X::Array{Int, 2} is obtained from a call to mono_grlex!(X, m).

When Ki==1 or Ki=0 return the index of all interactions.

    XC(1) XC(2) XC(3)  Degree
  +------------------------
1 |  0     0     0        0
2 |  0     0     1        1
3 |  0     1     0        1
4 |  1     0     0        1
5 |  0     0     2        2
6 |  0     1     1        2
7 |  0     2     0        2
8 |  1     0     1        2
9 |  1     1     0        2

10 | 2 0 0 2 11 | 0 0 3 3 12 | 0 1 2 3 13 | 0 2 1 3 14 | 0 3 0 3 15 | 1 0 2 3 16 | 1 1 1 3 17 | 1 2 0 3 18 | 2 0 1 3 19 | 2 1 0 3 20 | 3 0 0 3

So, for Ki = 2

Bool[true
 true
 true
 true
 true
 true
 true
 true
 true
 true
 true
 false
 false
 true
 false
 false
 false
 false
 false
 true]

Hermetic.include

Hermetic.integrate

Hermetic.integrate_polynomial — Function.

Calculate ∫ P(z) phi(z; 0, I) dz

Note: Gamma((1+j)/2) for j even is a gamma evaluated on half integer.

Int general Gamma(n/2) = (n-2)!!/(2^(n-1)/2)√π

Thus

gamma((1+f1[r])/2))/√π = (f1[r]-1)!!/2^(f1[r]/2)

Hermetic.integrate_polynomial_times_xn — Function.

Calculate ∫P(z) phi(z; 0, I) dz

Hermetic.mono_grlex! — Function.

mono_grlex!(X::Array{Int, 2}, m::Int)

Fills x::Array{Int, 2} with the grlx ordering for spacial dimension m.

Hermetic.mono_next_grlex! — Function.

mono_next_grlex!{T <: Int}(m::T, x::Array{T, 1})

Returns the next monomial in grlex order.

Hermetic.mono_rank_grlex — Function.

mono_rank_grlex(m, x)

returns the graded lexicographic ordering rank of a monomial in m dimensions.

Agrs:

m::Int: spatial dimension

x::Array{Int,1}: the m dimensional vector representing the monomial.

Returns:

The rank of the monomial.

Details:

The graded lexicographic ordering is used, over all M dimensional vectors with sum 0, then 1, then 2, ...

For example, if m = 3, the monomials are ordered in a sequence that begins

#  x[1]  x[2]  x[3]     SUM
  +------------------------
1 |  0     0     0        0
  |
2 |  0     0     1        1
3 |  0     1     0        1
4 |  1     0     0        1
  |
5 |  0     0     2        2
6 |  0     1     1        2
7 |  0     2     0        2
8 |  1     0     1        2
9 |  1     1     0        2

10 | 2 0 0 2 | 11 | 0 0 3 3 12 | 0 1 2 3 13 | 0 2 1 3 14 | 0 3 0 3 15 | 1 0 2 3 16 | 1 1 1 3 17 | 1 2 0 3 18 | 2 0 1 3 19 | 2 1 0 3 20 | 3 0 0 3 | 21 | 0 0 4 4 22 | 0 1 3 4 23 | 0 2 2 4 24 | 0 3 1 4 25 | 0 4 0 4 26 | 1 0 3 4 27 | 1 1 2 4 28 | 1 2 1 4 29 | 1 3 0 4 30 | 2 0 2 4 31 | 2 1 1 4 32 | 2 2 0 4 33 | 3 0 1 4 34 | 3 1 0 4 35 | 4 0 0 4

In the case of m = 1

   x[1]
  +----
1 |  0
2 |  1
3 |  2

The monomial (1,0,3) has rank 26, and we could determine this by the call rank = mono_rank_grlex (3, [1, 0, 3]);

The monomial (0,1,0) has rank 3, and we could determine this by the call rank = mono_rank_grlex ( 3, [0, 1, 0]);

Example:

mono_rank_grlex(3, [1,0,3]) return 26.

First compares the total degree (sum of all exponents), and in case of a tie applies lexicographic order. This ordering is not only a well ordering, it also has the property that any monomial is preceded only by a finite number of other monomials; this is not the case for lexicographic order, where all (infinitely many) powers of x are less than y (that lexicographic order is nevertheless a well ordering is related to the impossibility to construct an infinite decreasing chain of monomials). In other words,

x^\alpha >{grlex} x^{\beta} if deg x^\alpha > deg x^\beta or if deg x^\alpha > deg x^\beta if x^\alpha >{lex} x^\beta

where in the lexycographic order the power of the first variable is used to determine the order, with powers of the second variable only looked at when the first variable appears to the same power in two monomials.

Hermetic.mono_unrank_grlex — Function.

mono_unrank_grlex{T <: Int}(m::T, rank::T)

Computes the composition of given grlex rank.

Args:

- `m` the spatial dimension of the product poly
- `r` the rank of the composition

Output: - f the composition of the given rank

Example:

mono_unrank_grlex(3, 26) returns [1,0,3] f = Array{Int}(3, 1) `monounrankgrlex!(f, 3, 26) println(f)

Hermetic.mono_unrank_grlex!

Hermetic.mono_value — Function.

mono_value!(v, x, λ)

Evaluates a monomial.

Args:

x::Array{Float,2} the coordinates of the evaluation points.

λ::Array{Float,1} the exponents of the monomial (m×1)

Return the evaluated monomial v::Array{Float,1}.

Hermetic.poly_order

Hermetic.polynomial_add — Function.

polynomial_add(o1, c1, e1, o2, c2, e2)

Adds two polynomial

Args:

  - o1::Int the "order" of polynomial 1.

  - c1::Array{Float64,1}, the `O1×1` coefficients of polynomial 1.

  - e1::Array{Float64,1}, the `O1×1`, the indices of the exponents
of polynomial 1.

  - o2::Int the "order" of polynomial 2.

  - c2::Array{Float64,1}, the `O2×1` coefficients of polynomial 2

  - e1::Array{Float64,1}, the `O2×1`, the indices of the exponents
of polynomial 2.

Hermetic.polynomial_compress — Function.

polynomial_compress( o, c, e )

Args:

o the order of the polynomial

c the coefficients of the polynomial

e the indices of the exponents of the polynomial.

Output:

o the order of the compressed polynomial

c the coefficients of the compressed polynomial

e the indices of the exponents of the compressed polynomial.

Hermetic.polynomial_dif — Function.

polynomial_dif

Differentiates a polynomial

dif::Array{T}(m) indicates the number of differentiations in each component

Hermetic.polynomial_mul — Function.

polynomial_mul(m, o1, c1, e1, o2, c2, e2)

Multiply two polynomials

Args:

  - m::Int the spatial dimension of the product polynomial

  - o1::Int the "order" of polynomial 1.

  - c1::Array{Float64,1}, the `O1×1` coefficients of polynomial 1.

  - e1::Array{Float64,1}, the `O1×1`, the indices of the exponents
of polynomial 1.

  - o2::Int the "order" of polynomial 2.

  - c2::Array{Float64,1}, the `O2×1` coefficients of polynomial 2

  - e1::Array{Float64,1}, the `O2×1`, the indices of the exponents
of polynomial 2.

Hermetic.polynomial_mul_unc

Hermetic.polynomial_pow2

Hermetic.polynomial_print

Hermetic.polynomial_print_hermite

Hermetic.polynomial_scale — Function.

polynomial_scale{T <: AbstractFloat, F <: Int}(s, m::F, o::F, c::Array{T, 1}, e::Array{F, 1})

Scales a polynomial.

Args: - s the scale factor - m the spatial dimension - o the order of the polynomial - c the coefficients - e the exponent

Output: - o - c - e

Hermetic.polynomial_sort! — Function.

polynomial_sort! ( c, e )

sorts the information in a polynomial.

Details:

The coefficients c and exponents e are rearranged so that the elements of e are in ascending order.

c::Array{Float64,1} the coefficients of the polynomial.

e::Array{Int,1} the indices of the exponents of the polynomial.

Output:

c::Array{Float,1} the coefficients of the sorted polynomial.

e::Array{Int,1} the indices of the exponents of the sorted polynomial.

Hermetic.polynomial_value — Function.

polynomial_value

Input, int M, the spatial dimension.

Input, int O, the "order" of the polynomial.

Input, double C[O], the coefficients of the polynomial.

Input, int E[O], the indices of the exponents
of the polynomial.

Input, double X[NX, M], the coordinates of the evaluation points.

Hermetic.polyval

Hermetic.scale

Hermetic.setcoef!