Docstrings

LimitedLDLFactorizations.LimitedLDLFactorizations — Module.

A Pure Julia Version of LLDL. A left-looking implementation of the sparse LDL factorization of a symmetric matrix with the possibility to compute a limited-memory incomplete factorization.

Dominique Orban <dominique.orban@gmail.com> Montreal, April 2015, December 2015, July 2017.

This code is strongly inspired by Lin and Moré's ICFS [1,2]. The modified version is described in [3,4].

References

[1] C.-J. Lin and J. J. Moré. Incomplete Cholesky factorizations with limited memory. SIAM Journal on Scientific Computing, 21(1):24–45, 1999. [2] http://www.mcs.anl.gov/~more/icfs [3] D. Orban. Limited-Memory LDLT Factorization of Symmetric Quasi-Definite Matrices with Application to Constrained Optimization. Numerical Algorithms 70(1):9–41, 2015. DOI 10.1007/s11075-014-9933-x [4] https://github.com/optimizers/lldl

LimitedLDLFactorizations.abspermute! — Function.

Permute the elements of keys in place so that

abs(x[keys[i]]) ≤ abs(x[keys[k]])  for i = 1, ..., k
abs(x[keys[k]]) ≤ abs(x[keys[i]])  for i = k, ..., n,

where n is the length of keys. The length of x should be at least n. Only keys is modified.

From the MINPACK2 function dsel2 by Kastak, Lin and Moré.

LimitedLDLFactorizations.attempt_lldl!

LimitedLDLFactorizations.eval

LimitedLDLFactorizations.include

LimitedLDLFactorizations.lldl — Function.

lldl(A)

Compute the limited-memory LDLᵀ factorization of A without pivoting.

Arguments

A::SparseMatrixCSC{Tv,Ti}: matrix to factorize (its strict lower triangle and diagonal will be extracted)

Keyword arguments

memory::Int=0: extra amount of memory to allocate for the incomplete factor L. The total memory allocated is nnz(T) + n * memory, where T is the strict lower triangle of A and n is the size of A.

α::Tv=Tv(0): initial value of the shift in case the incomplete LDLᵀ factorization of A is found to not exist. The shift will be gradually increased from this initial value until success.

droptol::Tv=Tv(0): to further sparsify L, all elements with magnitude smaller than droptol are dropped.