Docstrings

SeriesAccelerators.SeriesAccelerators

SeriesAccelerators._memoise

SeriesAccelerators._seriesaccelerator

SeriesAccelerators._shanks

SeriesAccelerators._vanwijngaarden

SeriesAccelerators.check_convergence — Function.

Check the convergence of series by comparing the new value to the value of the series so far

SeriesAccelerators.default_atol

SeriesAccelerators.default_initial_iterate

SeriesAccelerators.default_recursion

SeriesAccelerators.default_rtol

SeriesAccelerators.default_sum_limit

SeriesAccelerators.eval

SeriesAccelerators.include

SeriesAccelerators.shanks — Function.

shanks(series, recursion::Int=1, sum_limit::U=1_000_000; rtol=sqrt(eps()), atol=eps())

Shanks transformation series accelerator. https://en.wikipedia.org/wiki/Shanks_transformation

Arguments: series (optional, Function) : a function that accepts an argument, n::Int, and returns the nth value in the series sum_limit (optional, Int) : the value at which to stop the series (default is 1,000,000) rtol (optional, Number) : relative stopping tolerance atol (optional, Number) : absolute stopping tolerance

SeriesAccelerators.vanwijngaarden — Function.

vanwijngaarden(series, recursion::Int=1, sum_limit::U=1_000_000; rtol=sqrt(eps()), atol=eps())

van Wijngaarden transformation series accelerator. https://en.wikipedia.org/wiki/VanWijngaardentransformation

Arguments: series (optional, Function) : a function that accepts an argument, n::Int, and returns the nth value in the series sum_limit (optional, Int) : the value at which to stop the series (default is 1,000,000) rtol (optional, Number) : relative stopping tolerance atol (optional, Number) : absolute stopping tolerance