Docstrings

StrRegex.@R_str

StrRegex.Binary_Regex_CSEs

StrRegex.DEFAULT_COMPILER_OPTS

StrRegex.DEFAULT_MATCH_OPTS

StrRegex.JIT_STACK_MAX_SIZE

StrRegex.JIT_STACK_START_SIZE

StrRegex.MATCH_CONTEXT

StrRegex.MS_Str

StrRegex.MS_String

StrRegex.MatchTab

StrRegex.PCRE

StrRegex.RegexStr

StrRegex.RegexStrMatch

StrRegex.RegexStrMatchIterator

StrRegex.RegexTypes

StrRegex.Regex_CSEs

StrRegex.StrRegex — Module.

Regex functions for Str strings

Copyright 2018 Gandalf Software, Inc., Scott P. Jones, and other contributors to the Julia language Licensed under MIT License, see LICENSE.md Based in part on julia/base/regex.jl and julia/base/pcre.jl

StrRegex._UCP

StrRegex._UTF

StrRegex._VALID

StrRegex.__api__

StrRegex.__find

StrRegex.__init__

StrRegex.__tmp_api__

StrRegex._add_compile_options

StrRegex._check_compile

StrRegex._check_match

StrRegex._clear_opts

StrRegex._finalize!

StrRegex._find

StrRegex._match

StrRegex._not_found

StrRegex._occurs_in

StrRegex._update_match

StrRegex._update_table

StrRegex._write_capture

StrRegex.check_compile

StrRegex.cmp_all — Function.

Precompile all of the common cases needed for String, UTF8Str, UTF16Str, and UniStr

StrRegex.codeunit_index

StrRegex.comp_add

StrRegex.comp_exec

StrRegex.contextlock

StrRegex.deb

StrRegex.empty_table

StrRegex.eval

StrRegex.exec

StrRegex.fin

StrRegex.finalize!

StrRegex.get_comp_ind

StrRegex.get_match_context — Function.

Get a thread-specific match context

StrRegex.get_ovec

StrRegex.get_range

StrRegex.get_regex

StrRegex.include

StrRegex.match_add

StrRegex.match_context

StrRegex.md_free

StrRegex.opt_index

StrRegex.outtab

StrRegex.regex_type_error

StrRegex.tabtype