Docstrings

NetworkLayout.Buchheim — Module.

Using the algorithm proposed in the paper, "Improving Walker's Algorithm to Run in Linear Time" by Christoph Buchheim, Michael Junger, Sebastian Leipert (http://dirk.jivas.de/papers/buchheim02improving.pdf)

Arguments tree Adjacency List that represents the given tree

Returns positions co-ordinates of the layout

NetworkLayout.Circular — Module.

This function wrap from NetworkX Position nodes on a circle. Parameters adj_matrix a graph Returns locsx, locsy Locations of the nodes. Can be any units you want, but will be normalized and centered anyway Examples

julia> g = simple_house_graph()
julia> locs_x, locs_y = circular_layout(g)

NetworkLayout.NetworkLayout

NetworkLayout.SFDP — Module.

Using the Spring-Electric model suggested by Yifan Hu (http://yifanhu.net/PUB/graphdrawsmall.pdf) Forces are calculated as : fattr(i,j) = ||xi - xj||^2 / K , i<->j frepln(i,j) = -CK^2 / ||xi - xj|| , i!=j Arguments : adj_matrix Sparse/Full Adjacency matrix of the graph tol Tolerance distance - Minimum distance between 2 nodes C, K Constants that help scale the layout Output : positions Co-ordinates for the nodes

NetworkLayout.Shell — Module.

This function is copy from IainNZ's GraphLayout.jl Position nodes in concentric circles. Parameters adj_matrix a graph nlist Vector of Vector, Vector of node Vector for each shell. Examples

julia> g = graphfamous("karate")
julia> nlist = Array(Vector{Int}, 2)
julia> nlist[1] = [1:5]
julia> nlist[2] = [6:num_vertiecs(g)]
julia> locs_x, locs_y = shell_layout(g, nlist)

NetworkLayout.Spectral

NetworkLayout.Spring — Module.

Use the spring/repulsion model of Fruchterman and Reingold (1991):
    Attractive force:  f_a(d) =  d^2 / k
    Repulsive force:  f_r(d) = -k^2 / d
where d is distance between two vertices and the optimal distance
between vertices k is defined as C * sqrt( area / num_vertices )
where C is a parameter we can adjust

Arguments:
adj_matrix    Adjacency matrix of some type. Non-zero of the eltype
              of the matrix is used to determine if a link exists,
              but currently no sense of magnitude
C             Constant to fiddle with density of resulting layout
iterations    Number of iterations we apply the forces
initialtemp   Initial "temperature", controls movement per iteration

NetworkLayout.Stress — Module.

Compute graph layout using stress majorization

Inputs:

δ: Matrix of pairwise distances
p: Dimension of embedding (default: 2)
weights: Matrix of weights. If not specified, defaults to
       weights[i,j] = δ[i,j]^-2 if δ[i,j] is nonzero, or 0 otherwise
X0: Initial guess for the layout. Coordinates are given in rows.
    If not specified, default to random matrix of Gaussians

Additional optional keyword arguments control the convergence of the algorithm and the additional output as requested:

iterations:   Maximum number of iterations. Default: 400size(X0, 1)^2
abstols:      Absolute tolerance for convergence of stress.
              The iterations terminate if the difference between two
              successive stresses is less than abstol.
              Default: √(eps(eltype(X0))
reltols:      Relative tolerance for convergence of stress.
              The iterations terminate if the difference between two
              successive stresses relative to the current stress is less than
              reltol. Default: √(eps(eltype(X0))
abstolx:      Absolute tolerance for convergence of layout.
              The iterations terminate if the Frobenius norm of two successive
              layouts is less than abstolx. Default: √(eps(eltype(X0))

Output:

The final layout positions.

Reference:

The main equation to solve is (8) of:

@incollection{
    author = {Emden R Gansner and Yehuda Koren and Stephen North},
    title = {Graph Drawing by Stress Majorization}
    year={2005},
    isbn={978-3-540-24528-5},
    booktitle={Graph Drawing},
    seriesvolume={3383},
    series={Lecture Notes in Computer Science},
    editor={Pach, J'anos},
    doi={10.1007/978-3-540-31843-9_25},
    publisher={Springer Berlin Heidelberg},
    pages={239--250},
}

NetworkLayout.eval

NetworkLayout.include