Docstrings

Munkres.Location

Munkres.ModifiedCost

Munkres.Munkres

Munkres.PrimeMark

Munkres.StarMark

Munkres.augment_path!

Munkres.erase_primes!

Munkres.eval

Munkres.find_prime_in_row

Munkres.find_smallest_uncovered

Munkres.find_star_in_column

Munkres.include

Munkres.iszero

Munkres.munkres — Function.

Munkres (Hungarian) Algorithm for optimal assignment, i.e. given an NxM cost matrix for assigning N workers to M jobs, what is the optimal allocation of tasks to workers such that one one task is assigned to one worker.

Input

cost_matrix - an NxM real valued matrix of the cost of assigning job m to worker N

Output

p - A vector of length N of assignments, where j = p[i] assigns job j to worker i.

Munkres.step_five!

Munkres.step_four!

Munkres.step_one!

Munkres.step_six!

Munkres.step_three!

Munkres.step_two!

Munkres.valid