Docstrings

MetaImageFormat.MetaImageFormat

MetaImageFormat.UnknownColor — Type.

UnknownColor{T,N}

An unknown Color. This type gets returned when one of the "kind" settings is "3-color" or "4-color".

MetaImageFormat.VTuple

MetaImageFormat.arraytype — Function.

arraytype(filename)

Parse MetaImage header and call arraytype!(header). See arraytype! for information about the return values.

MetaImageFormat.arraytype! — Function.

arraytype!(header) -> T, axs, perm, need_bswap

Analyze the header dictionary to extract the element-type T, size or axes information axs, the permutation perm (if any) that julia should use for "wrapping" the read data, and a boolean need_bswap indicating whether the data need to be byte-swapped (to account for differences in endianness). T includes any color information (in which case a dimension of the array will be "consumed"). axs will be a Dims-tuple in simple cases, or an Axes tuple (from AxisArrays.jl) if dimensions are labeled or have their spatial information (pixelspacing, spacedirections, etc) specified. perm is the permutation needed to move the color data to the first dimension, or an empty tuple if no permutation is required.

This function may modify the header dictionary (the reason for the ! in the name), so make a copy first if necessary.

MetaImageFormat.axes2space

MetaImageFormat.checked_size

MetaImageFormat.chksize

MetaImageFormat.colorant_eltype — Function.

colorant_eltype(C, T) -> Tc

Return a valid "inner" element type Tc for colorant type C. When T != Tc, values must be "converted" before they can be interpreted as type C.

MetaImageFormat.eval

MetaImageFormat.find_datafile

MetaImageFormat.fixedtype — Function.

fixedtype(Traw, header) -> Tu

Attempt to interpret type Traw in terms of FixedPoint numbers. If Traw cannot be interpreted as Normed, Tu = Traw.

MetaImageFormat.fixedtype_max

MetaImageFormat.get_axes

MetaImageFormat.get_size

MetaImageFormat.include

MetaImageFormat.load

MetaImageFormat.meta_format

MetaImageFormat.meta_parse

MetaImageFormat.mymsb

MetaImageFormat.outer_eltype! — Function.

outer_eltype!(header, Traw) -> T, nd, perm

Extract the julia array eltype T, the number of dimensions nd excluding color/complex/vector/matrix element data, and any permutation needed to put the eltype dimension first. Any dimensions in the header corresponding to color (or if "kind" is set to one of the vector types) will be "consumed" upon exit. Traw is the element type as determined by raw_eltype.

See also: raw_eltype.

MetaImageFormat.parse_header — Function.

header = parse_header(io)

Parse the MetaImage header (an .mhd file). io should be positioned at the beginning of the header file.

See also: write_header.

MetaImageFormat.parse_type

MetaImageFormat.raw_eltype — Function.

raw_eltype(header) -> Traw, need_bswap
raw_eltype(::Type{T}) -> Traw

Get the "basic" element type of the data, e.g., UInt16 or Float32.

This function does not try to determine whether the image is color (Traw does not contain any color information), nor does it try to interpret Traw as a Normed type.

See also: outer_eltype!, fixedtype.

MetaImageFormat.space2axes

MetaImageFormat.spacings

MetaImageFormat.stream2name

MetaImageFormat.string2type

MetaImageFormat.type2string