Docstrings

AMQPClient.@debug

AMQPClient.AMQPClient

AMQPClient.AMQPClientException

AMQPClient.AMQPProtocolException

AMQPClient.AMQP_DEFAULT_PORT

AMQPClient.AMQP_VERSION

AMQPClient.AUTH_PROVIDERS

AMQPClient.AbstractChannel

AMQPClient.CLASSMETHODNAME_MAP

AMQPClient.CLASSNAME_MAP

AMQPClient.CLASS_MAP

AMQPClient.CLIENT_IDENTIFICATION

AMQPClient.CONN_MAX_QUEUED

AMQPClient.CONN_STATE_CLOSED

AMQPClient.CONN_STATE_CLOSING

AMQPClient.CONN_STATE_OPEN

AMQPClient.CONN_STATE_OPENING

AMQPClient.ClassSpec

AMQPClient.CloseReason

AMQPClient.Connection

AMQPClient.ContentWeight

AMQPClient.DEBUG

AMQPClient.DEFAULT_AUTH_PARAMS

AMQPClient.DEFAULT_CHANNEL

AMQPClient.DEFAULT_CHANNELMAX

AMQPClient.DEFAULT_CONNECT_TIMEOUT

AMQPClient.DEFAULT_TIMEOUT

AMQPClient.EXCHANGE_TYPE_DIRECT

AMQPClient.EXCHANGE_TYPE_FANOUT

AMQPClient.EXCHANGE_TYPE_HEADERS

AMQPClient.EXCHANGE_TYPE_TOPIC

AMQPClient.FieldIndicatorMap

AMQPClient.FieldValueIndicatorMap

AMQPClient.FrameBody

AMQPClient.FrameEnd

AMQPClient.FrameHeader

AMQPClient.FrameHeartbeat

AMQPClient.FrameMethod

AMQPClient.FrameMinSize

AMQPClient.GET_EMPTY_RESP

AMQPClient.HardErrorChannelError

AMQPClient.HardErrorCommandInvalid

AMQPClient.HardErrorConnectionForced

AMQPClient.HardErrorFrameError

AMQPClient.HardErrorInternalError

AMQPClient.HardErrorInvalidPath

AMQPClient.HardErrorNotAllowed

AMQPClient.HardErrorNotImplemented

AMQPClient.HardErrorResourceError

AMQPClient.HardErrorSyntaxError

AMQPClient.HardErrorUnexpectedFrame

AMQPClient.HeartBeat

AMQPClient.LiteralAMQP

AMQPClient.Message

AMQPClient.MessageChannel

AMQPClient.MessageConsumer

AMQPClient.MethodSpec

AMQPClient.NON_PERSISTENT

AMQPClient.PERSISTENT

AMQPClient.PROPERTIES

AMQPClient.PropertySpec

AMQPClient.ProtocolHeader

AMQPClient.ProtocolId

AMQPClient.ProtocolVersion

AMQPClient.ReplySuccess

AMQPClient.SORTED_PROPERTIES

AMQPClient.SORTED_PROPERTY_NAMES

AMQPClient.SoftErrorAccessRefused

AMQPClient.SoftErrorContentTooLarge

AMQPClient.SoftErrorNoConsumers

AMQPClient.SoftErrorNotFound

AMQPClient.SoftErrorPreconditionFailed

AMQPClient.SoftErrorResourceLocked

AMQPClient.TAMQPBit

AMQPClient.TAMQPBodyPayload

AMQPClient.TAMQPBool

AMQPClient.TAMQPByteArray

AMQPClient.TAMQPChannel

AMQPClient.TAMQPClassId

AMQPClient.TAMQPConsumerTag

AMQPClient.TAMQPContent

AMQPClient.TAMQPContentBodyFrame

AMQPClient.TAMQPContentBodySize

AMQPClient.TAMQPContentClass

AMQPClient.TAMQPContentHeaderFrame

AMQPClient.TAMQPDecimalValue

AMQPClient.TAMQPDeliveryTag

AMQPClient.TAMQPDouble

AMQPClient.TAMQPExchangeName

AMQPClient.TAMQPFV

AMQPClient.TAMQPField

AMQPClient.TAMQPFieldArray

AMQPClient.TAMQPFieldName

AMQPClient.TAMQPFieldTable

AMQPClient.TAMQPFieldValue

AMQPClient.TAMQPFieldValuePair

AMQPClient.TAMQPFloat

AMQPClient.TAMQPFrameProperties

AMQPClient.TAMQPGenericFrame

AMQPClient.TAMQPHeaderPayload

AMQPClient.TAMQPHeartBeatFrame

AMQPClient.TAMQPLengthPrefixed

AMQPClient.TAMQPLongInt

AMQPClient.TAMQPLongLongInt

AMQPClient.TAMQPLongLongUInt

AMQPClient.TAMQPLongStr

AMQPClient.TAMQPLongUInt

AMQPClient.TAMQPMessageCount

AMQPClient.TAMQPMethod

AMQPClient.TAMQPMethodFrame

AMQPClient.TAMQPMethodId

AMQPClient.TAMQPMethodPayload

AMQPClient.TAMQPNoAck

AMQPClient.TAMQPNoLocal

AMQPClient.TAMQPNoWait

AMQPClient.TAMQPOctet

AMQPClient.TAMQPPath

AMQPClient.TAMQPPayloadSize

AMQPClient.TAMQPPeerProperties

AMQPClient.TAMQPPropertyFlags

AMQPClient.TAMQPQueueName

AMQPClient.TAMQPRedelivered

AMQPClient.TAMQPReplyCode

AMQPClient.TAMQPReplyText

AMQPClient.TAMQPScale

AMQPClient.TAMQPShortInt

AMQPClient.TAMQPShortShortInt

AMQPClient.TAMQPShortShortUInt

AMQPClient.TAMQPShortStr

AMQPClient.TAMQPShortUInt

AMQPClient.TAMQPTimeStamp

AMQPClient.UNUSED_CHANNEL

AMQPClient._on_ack

AMQPClient._on_basic_consume_cancel_ok

AMQPClient._on_close_ok

AMQPClient._on_queue_purge_delete_ok

AMQPClient._send_close

AMQPClient._send_close_ok

AMQPClient._send_exchange_bind_unbind

AMQPClient._send_tx

AMQPClient._tx

AMQPClient._wait_resp

AMQPClient.as_fval

AMQPClient.auth_resp_amqplain

AMQPClient.basic_ack

AMQPClient.basic_cancel — Function.

Cancels a consumer.

This does not affect already delivered messages, but it does mean the server will not send any more messages for that consumer. The client may receive an arbitrary number of messages in between sending the cancel method and receiving the cancel­ok reply.

AMQPClient.basic_consume — Function.

Start a queue consumer.

queue: queue name consumertag: id of the consumer, server generates a unique tag if this is empty nolocal: do not deliver own messages no_ack: no acknowledgment needed, server automatically and silently acknowledges delivery (speed at the cost of reliability) exclusive: request exclusive access (only this consumer can access the queue) nowait: do not send a reply method

AMQPClient.basic_get

AMQPClient.basic_publish — Function.

Publish a message

This method publishes a message to a specific exchange. The message will be routed to queues as defined by the exchange configuration and distributed to any active consumers when the transaction, if any, is committed.

AMQPClient.basic_qos

AMQPClient.basic_recover

AMQPClient.basic_reject

AMQPClient.channel

AMQPClient.channel_message_consumer

AMQPClient.channel_receiver

AMQPClient.clear_handlers

AMQPClient.close

AMQPClient.confirm_select

AMQPClient.connection

AMQPClient.connection_heartbeater

AMQPClient.connection_processor

AMQPClient.connection_receiver

AMQPClient.connection_sender

AMQPClient.convert

AMQPClient.default_exchange_name

AMQPClient.displayname

AMQPClient.eval

AMQPClient.exchange_bind

AMQPClient.exchange_declare

AMQPClient.exchange_delete

AMQPClient.exchange_unbind

AMQPClient.find_unused_channel

AMQPClient.frame_key

AMQPClient.get_property

AMQPClient.handle

AMQPClient.include

AMQPClient.is_method — Function.

Validate if the method frame is for the given class and method.

AMQPClient.make_classmethod_map

AMQPClient.method

AMQPClient.method_key

AMQPClient.method_name

AMQPClient.methodargs

AMQPClient.on_basic_cancel_ok

AMQPClient.on_basic_consume_ok

AMQPClient.on_basic_get_empty_or_ok

AMQPClient.on_basic_qos_ok

AMQPClient.on_basic_recover_ok

AMQPClient.on_channel_close

AMQPClient.on_channel_close_ok

AMQPClient.on_channel_flow

AMQPClient.on_channel_message_in

AMQPClient.on_channel_open_ok

AMQPClient.on_confirm_select_ok

AMQPClient.on_connection_close

AMQPClient.on_connection_close_ok

AMQPClient.on_connection_heartbeat

AMQPClient.on_connection_open_ok

AMQPClient.on_connection_start

AMQPClient.on_connection_tune

AMQPClient.on_exchange_bind_ok

AMQPClient.on_exchange_declare_ok

AMQPClient.on_exchange_delete_ok

AMQPClient.on_exchange_unbind_ok

AMQPClient.on_queue_bind_ok

AMQPClient.on_queue_declare_ok

AMQPClient.on_queue_delete_ok

AMQPClient.on_queue_purge_ok

AMQPClient.on_queue_unbind_ok

AMQPClient.on_tx_commit_ok

AMQPClient.on_tx_rollback_ok

AMQPClient.on_tx_select_ok

AMQPClient.on_unexpected_message

AMQPClient.queue_bind

AMQPClient.queue_declare — Function.

Declare a queue (or query an existing queue). Returns a tuple: (boolean success/failure, queue name, message count, consumer count)

AMQPClient.queue_delete — Function.

Delete a queue. Returns a tuple: (boolean success/failure, message count)

AMQPClient.queue_purge — Function.

Purge messages from a queue. Returns a tuple: (boolean success/failure, message count)

AMQPClient.queue_unbind

Base.read — Function.

Read a generic frame. All frames have the following wire format:

0 1 3 7 size+7 size+8 +–––+––––-+––––-+ +––––––-+ +–––––-+ | type | channel | size | | payload | | frame-end | +–––+––––-+––––-+ +––––––-+ +–––––-+ octet short long 'size' octets octet

AMQPClient.read!

AMQPClient.send

AMQPClient.send_basic_ack

AMQPClient.send_basic_cancel

AMQPClient.send_basic_consume

AMQPClient.send_basic_get

AMQPClient.send_basic_publish

AMQPClient.send_basic_qos

AMQPClient.send_basic_recover

AMQPClient.send_basic_reject

AMQPClient.send_channel_close

AMQPClient.send_channel_close_ok

AMQPClient.send_channel_flow

AMQPClient.send_channel_open

AMQPClient.send_connection_close

AMQPClient.send_connection_close_ok

AMQPClient.send_connection_heartbeat

AMQPClient.send_connection_open

AMQPClient.send_connection_start_ok

AMQPClient.send_connection_tune_ok

AMQPClient.send_exchange_bind

AMQPClient.send_exchange_declare

AMQPClient.send_exchange_delete

AMQPClient.send_exchange_unbind

AMQPClient.send_queue_bind

AMQPClient.send_queue_declare

AMQPClient.send_queue_delete

AMQPClient.send_queue_purge

AMQPClient.send_queue_unbind

AMQPClient.send_tx_commit

AMQPClient.send_tx_rollback

AMQPClient.send_tx_select

AMQPClient.set_properties

AMQPClient.show

AMQPClient.sock

AMQPClient.tx_commit

AMQPClient.tx_rollback

AMQPClient.tx_select

AMQPClient.wait_for_state