Docstrings

RegressionTables.RegressionTable

RegressionTables.RegressionTables

RegressionTables.RenderSettings

RegressionTables.asciiOutput

RegressionTables.columns

RegressionTables.default_ascii_estim_decoration

RegressionTables.default_latex_estim_decoration

RegressionTables.eval

RegressionTables.htmlOutput

RegressionTables.include

RegressionTables.isFERegressionResult

RegressionTables.isIVRegressionResult

RegressionTables.isOLSRegressionResult

RegressionTables.latexOutput

RegressionTables.regtable — Function.

Produces a publication-quality regression table, similar to Stata's esttab and R's stargazer.

Arguments

rr::AbstractRegressionResult... are the AbstractRegressionResults from FixedEffectModels.jl that should be printed. Only required argument.

regressors is a Vector of regressor names (Strings) that should be shown, in that order. Defaults to an empty vector, in which case all regressors will be shown.

fixedeffects is a Vector of FE names (Strings) that should be shown, in that order. Defaults to an empty vector, in which case all FE's will be shown.

labels is a Dict that contains displayed labels for variables (strings) and other text in the table. If no label for a variable is found, it default to variable names. See documentation for special values.

estimformat is a String that describes the format of the estimate. Defaults to "%0.3f".

estim_decoration is a Function that takes the formatted string and the p-value, and applies decorations (such as the beloved stars). Defaults to (* p<0.05, ** p<0.01, *** p<0.001).

statisticformat is a String that describes the format of the number below the estimate (se/t). Defaults to "%0.4f".

below_statistic is a Symbol that describes a statistic that should be shown below each point estimate. Recognized values are :blank, :se, and :tstat. Defaults to :se.

below_decoration is a Function that takes the formatted statistic string, and applies a decorations. Defaults to round parentheses.

regression_statistics is a Vector of Symbols that describe statistics to be shown at the bottom of the table. Recognized symbols are :nobs, :r2, :r2_a, :r2_within, :f, :p, :f_kp, :p_kp, and :dof. Defaults to [:nobs, :r2].

number_regressions is a Bool that governs whether regressions should be numbered. Defaults to true.

number_regressions_decoration is a Function that governs the decorations to the regression numbers. Defaults to s -> "($s)".

print_fe_section is a Bool that governs whether a section on fixed effects should be shown. Defaults to true.

print_estimator_section is a Bool that governs whether to print a section on which estimator (OLS/IV/NL) is used. Defaults to true.

standardize_coef is a Bool that governs whether the table should show standardized coefficients. Note that this only works with DataFrameRegressionModels, and that only coefficient estimates and the below_statistic are being standardized (i.e. the R^2 etc still pertain to the non-standardized regression).

out_buffer is an IOBuffer that the output gets sent to (unless an output file is specified, in which case the output is only sent to the file).

renderSettings::RenderSettings is a RenderSettings composite type that governs how the table should be rendered. Standard supported types are ASCII (via asciiOutput(outfile::String)) and LaTeX (via latexOutput(outfile::String)). If no argument to these two functions are given, the output is sent to STDOUT. Defaults to ASCII with STDOUT.

Details

A typical use is to pass a number of AbstractRegressionResults to the function, along with a RenderSettings object.

regtable(regressionResult1, regressionResult2; renderSettings = asciiOutput())

Pass a string to the functions that create a RenderSettings to divert output to a file. For example, using LaTeX output,

regtable(regressionResult1, regressionResult2; renderSettings = latexOutput("myoutfile.tex"))

See the full argument list for details.

Examples

using RegressionTables, DataFrames, RDatasets, FixedEffectModels
df = dataset("datasets", "iris")
df[:SpeciesDummy] = pool(df[:Species])
df[:isSmall] = pool(df[:SepalWidth] .< 2.9)
rr1 = reg(df, @model(SepalLength ~ SepalWidth))
rr2 = reg(df, @model(SepalLength ~ SepalWidth + PetalLength   , fe = SpeciesDummy))
rr3 = reg(df, @model(SepalLength ~ SepalWidth + PetalLength + PetalWidth  , fe = SpeciesDummy  + isSmall))
rr4 = reg(df, @model(SepalWidth ~ SepalLength + PetalLength + PetalWidth  , fe = SpeciesDummy))
rr5 = reg(df, @model(SepalWidth ~ SepalLength + (PetalLength ~ PetalWidth)  , fe = SpeciesDummy))
# default
regtable(rr1,rr2,rr3,rr4; renderSettings = asciiOutput())
# display of statistics below estimates
regtable(rr1,rr2,rr3,rr4; renderSettings = asciiOutput(), below_statistic = :blank)
regtable(rr1,rr2,rr3,rr4; renderSettings = asciiOutput(), below_decoration = s -> "[$(s)]")
# ordering of regressors, leaving out regressors
regtable(rr1,rr2,rr3,rr4; renderSettings = asciiOutput(), regressors = ["SepalLength";"PetalWidth";"SepalWidth"])
# format of the estimates
regtable(rr1,rr2,rr3,rr4; renderSettings = asciiOutput(), estimformat = "%02.5f")
# replace some variable names by other strings
regtable(rr1,rr2,rr3; renderSettings = asciiOutput(), labels = Dict("SepalLength" => "My dependent variable: SepalLength", "PetalLength" => "Length of Petal", "PetalWidth" => "Width of Petal", "(Intercept)" => "Const." , "isSmall" => "isSmall Dummies", "SpeciesDummy" => "Species Dummies"))
# do not print the FE block
regtable(rr1,rr2,rr3,rr4; renderSettings = asciiOutput(), print_fe_section = false)
# re-order fixed effects
regtable(rr1,rr2,rr3,rr4; renderSettings = asciiOutput(), fixedeffects = ["isSmall", "SpeciesDummy"])
# change the yes/no labels in the fixed effect section, and statistics labels
regtable(rr1,rr2,rr3,rr4; renderSettings = asciiOutput(), labels = Dict("__LABEL_FE_YES__" => "Mhm.", "__LABEL_FE_NO__" => "Nope.", "__LABEL_STATISTIC_N__" => "Number of observations", "__LABEL_STATISTIC_R2__" => "R Squared"))
# full set of available statistics
regtable(rr1,rr2,rr3,rr5; renderSettings = asciiOutput(), regression_statistics = [:nobs, :r2, :r2_a, :r2_within, :f, :p, :f_kp, :p_kp, :dof])
# LaTeX output
regtable(rr1,rr2,rr3,rr4; renderSettings = latexOutput())
# LaTeX output to file
regtable(rr1,rr2,rr3,rr4; renderSettings = latexOutput("myoutfile.tex"))

RegressionTables.render