Docstrings

PolynomialRoots.FRAC_JUMPS

PolynomialRoots.FRAC_JUMP_EVERY

PolynomialRoots.FRAC_JUMP_LEN

PolynomialRoots.MAX_ITERS

PolynomialRoots.PolynomialRoots

PolynomialRoots.divide_poly_1

PolynomialRoots.eval

PolynomialRoots.eval_poly_der

PolynomialRoots.eval_poly_der2

PolynomialRoots.eval_poly_der2_ek

PolynomialRoots.eval_poly_der_ek

PolynomialRoots.evalpoly

PolynomialRoots.find_2_closest_from_5

PolynomialRoots.include

PolynomialRoots.laguerre

PolynomialRoots.laguerre2newton

PolynomialRoots.newton_spec

PolynomialRoots.roots — Function.

roots(polynomial[, roots]; polish=true, epsilon=1e-20) -> roots

Find all the roots of polynomial, of any degree.

Mandatory argument:

polynomial: vector of coefficients (type Number) of the polynomial of which to find the roots, from the lowest coefficient to the highest one

Optional argument:

roots: vector of initial guess roots. If you have a very rough idea where some of the roots can be, this vector is used as starting value for Laguerre's method

Optional keywords:

polish (Bool): if set to true, after all roots have been found by dividing original polynomial by each root found, all roots will be polished using full polynomial. Default is false

epsilon (AbstractFloat): this is used to determine the stopping criterion described in Skowron & Gould paper. If not set, it defaults to machine precision of polynomial (and roots) argument(s). This is not the precision with which the roots will be calculated

Function root5 is specialized for polynomials of degree 5.

PolynomialRoots.roots!

PolynomialRoots.roots5 — Function.

roots5(polynomial[, roots]; epsilon=1e-20) -> roots

Find all the roots of polynomial, of degree 5 only.

Mandatory argument:

polynomial: vector of 6 coefficients (type Number) of the polynomial of which to find the roots, from the lowest coefficient to the highest one

Optional argument:

roots: vector of initial guess roots (of length 5). If you have a very rough idea where some of the roots can be, this vector is used as starting value for Laguerre's method and the provided roots will be only polished

Optional keyword:

epsilon (AbstractFloat): this is used to determine the stopping criterion described in Skowron & Gould paper. If not set, it defaults to machine precision of polynomial (and roots) argument(s). This is not the precision with which the roots will be calculated

Function roots can be used to find roots of polynomials of any degree.

PolynomialRoots.roots5!

PolynomialRoots.solve_cubic_eq

PolynomialRoots.solve_quadratic_eq

PolynomialRoots.sort_5_points_by_separation!

PolynomialRoots.sort_5_points_by_separation_i

PolynomialRoots.third