Docstrings

TickTock.TickTock — Module.

This module provides tick(), tock(), and tok() functions.

They're similar to the tic(), toc(), and toq() functions that you might find in MATLAB and similar software.

Don't use these for timing code execution! Julia provides much better facilities for measuring performance, ranging from the @time and @elapsed macros to packages such as BenchmarkTools.jl. (And remember, don't time Julia code running in global scope!)

This code used to live in Julia Base as the tic(), toc(), and toq() functions (in base/util.jl). They were deprecated in GitHub issue 17046.

TickTock.eval

TickTock.gettimers

TickTock.include

TickTock.laptimer — Function.

laptimer()

Print the current elapsed time, in canonical form, since the previous tick(), and continue counting.

TickTock.peektimer — Function.

peektimer()

Return the current elapsed seconds counted by the most recent timer, without stopping it.

TickTock.printcanonical

TickTock.showtimes

TickTock.tick — Function.

tick()

Start counting. The other functions are:

tock() ` stop counting, show total elapsed time in canonical form

tok() ` stop counting, return seconds

peektimer() continue counting, return elapsed seconds

laptimer() continue counting, show total elapsed time in canonical form

TickTock.tock — Function.

tock()

Print the elapsed time, in canonical form, of the most recent timer, then stop counting.

TickTock.tok — Function.

tok()

Return the current elapsed seconds counted by the most recent timer, then stop counting.