Docstrings

Krylov.@kaxpby!

Krylov.@kaxpy!

Krylov.@kdot

Krylov.@knrm2

Krylov.@kscal!

Krylov.Krylov

Krylov.KrylovStats — Type.

Abstract type for statistics returned by a solver

Krylov.LanczosStats

Krylov.SimpleStats — Type.

Type for statistics returned by non-Lanczos solvers

Krylov.SymmlqStats

Krylov.cg — Function.

The conjugate gradient method to solve the symmetric linear system Ax=b.

The method does not abort if A is not definite.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite.

Krylov.cg_lanczos — Function.

The Lanczos version of the conjugate gradient method to solve the symmetric linear system

Ax = b

The method does not abort if A is not definite.

Krylov.cg_lanczos_shift_seq — Function.

The Lanczos version of the conjugate gradient method to solve a family of shifted systems

(A + αI) x = b (α = α₁, α₂, ...)

The method does not abort if A + αI is not definite.

Krylov.cgls — Function.

Solve the regularized linear least-squares problem

minimize ‖b - Ax‖₂² + λ ‖x‖₂²

using the Conjugate Gradient (CG) method, where λ ≥ 0 is a regularization parameter. This method is equivalent to applying CG to the normal equations

(A'A + λI) x = A'b

but is more stable.

CGLS produces monotonic residuals ‖r‖₂ but not optimality residuals ‖A'r‖₂. It is formally equivalent to LSQR, though can be slightly less accurate, but simpler to implement.

Krylov.cgne — Function.

Solve the consistent linear system

Ax + √λs = b

using the Conjugate Gradient (CG) method, where λ ≥ 0 is a regularization parameter. This method is equivalent to applying CG to the normal equations of the second kind

(AA' + λI) y = b

but is more stable. When λ = 0, this method solves the minimum-norm problem

min ‖x‖₂ s.t. Ax = b.

When λ > 0, it solves the problem

min ‖(x,s)‖₂ s.t. Ax + √λs = b.

CGNE produces monotonic errors ‖x-x*‖₂ but not residuals ‖r‖₂. It is formally equivalent to CRAIG, though can be slightly less accurate, but simpler to implement. Only the x-part of the solution is returned.

A preconditioner M may be provided in the form of a linear operator.

Krylov.cgs — Function.

Solve the consistent linear system Ax = b using conjugate gradient squared algorithm.

From "Iterative Methods for Sparse Linear Systems (Y. Saad)" :

«The method is based on a polynomial variant of the conjugate gradients algorithm. Although related to the so-called bi-conjugate gradients (BCG) algorithm, it does not involve adjoint matrix-vector multiplications, and the expected convergence rate is about twice that of the BCG algorithm.

The Conjugate Gradient Squared algorithm works quite well in many cases. However, one difficulty is that, since the polynomials are squared, rounding errors tend to be more damaging than in the standard BCG algorithm. In particular, very high variations of the residual vectors often cause the residual norms computed to become inaccurate.

TFQMR and BICGSTAB were developed to remedy this difficulty.»

This implementation allows a right preconditioner M.

Krylov.cr — Function.

A truncated version of Stiefel’s Conjugate Residual method to solve the symmetric linear system Ax=b. The matrix A must be positive semi-definite.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite. In a linesearch context, 'linesearch' must be set to 'true'.

Krylov.craig — Function.

Find the least-norm solution of the consistent linear system

Ax + √λs = b

using the Golub-Kahan implementation of Craig's method, where λ ≥ 0 is a regularization parameter. This method is equivalent to CGNE but is more stable.

For a system in the form Ax = b, Craig's method is equivalent to applying CG to AA'y = b and recovering x = A'y. Note that y are the Lagrange multipliers of the least-norm problem

minimize ‖x‖ subject to Ax = b.

In this implementation, both the x and y-parts of the solution are returned.

Krylov.craigmr — Function.

Solve the consistent linear system

Ax + √λs = b

using the CRAIG-MR method, where λ ≥ 0 is a regularization parameter. This method is equivalent to applying the Conjugate Residuals method to the normal equations of the second kind

(AA' + λI) y = b

but is more stable. When λ = 0, this method solves the minimum-norm problem

min ‖x‖₂ s.t. x ∈ argmin ‖Ax - b‖₂.

When λ > 0, this method solves the problem

min ‖(x,s)‖₂ s.t. Ax + √λs = b.

CRAIGMR produces monotonic residuals ‖r‖₂. It is formally equivalent to CRMR, though can be slightly more accurate, and intricate to implement. Both the x- and y-parts of the solution are returned.

Krylov.crls — Function.

Solve the linear least-squares problem

minimize ‖b - Ax‖₂² + λ ‖x‖₂²

using the Conjugate Residuals (CR) method. This method is equivalent to applying MINRES to the normal equations

(A'A + λI) x = A'b.

This implementation recurs the residual r := b - Ax.

CRLS produces monotonic residuals ‖r‖₂ and optimality residuals ‖A'r‖₂. It is formally equivalent to LSMR, though can be slightly less accurate, but simpler to implement.

Krylov.crmr — Function.

Solve the consistent linear system

Ax + √λs = b

using the Conjugate Residual (CR) method, where λ ≥ 0 is a regularization parameter. This method is equivalent to applying CR to the normal equations of the second kind

(AA' + λI) y = b

but is more stable. When λ = 0, this method solves the minimum-norm problem

min ‖x‖₂ s.t. x ∈ argmin ‖Ax - b‖₂.

When λ > 0, this method solves the problem

min ‖(x,s)‖₂ s.t. Ax + √λs = b.

CGMR produces monotonic residuals ‖r‖₂. It is formally equivalent to CRAIG-MR, though can be slightly less accurate, but simpler to implement. Only the x-part of the solution is returned.

A preconditioner M may be provided in the form of a linear operator.

Krylov.diom — Function.

Solve the consistent linear system Ax = b using direct incomplete orthogonalization method.

DIOM is similar to CG with partial reorthogonalization.

An advantage of DIOM is that nonsymmetric or symmetric indefinite or both nonsymmetric and indefinite systems of linear equations can be handled by this single algorithm.

This implementation allows a right preconditioner M.

Krylov.dqgmres — Function.

Solve the consistent linear system Ax = b using DQGMRES method.

DQGMRES algorithm is based on the incomplete Arnoldi orthogonalization process and computes a sequence of approximate solutions with the quasi-minimal residual property.

This implementation allows a right preconditioner M.

Krylov.eval

Krylov.include

Krylov.krylov_axpby!

Krylov.krylov_axpy!

Krylov.krylov_dot

Krylov.krylov_norm2

Krylov.krylov_scal!

Krylov.lslq — Function.

lslq(A, b, λ=0.0)

Solve the regularized linear least-squares problem

minimize ‖b - Ax‖₂² + λ² ‖x‖₂²

using the LSLQ method, where λ ≥ 0 is a regularization parameter. LSLQ is formally equivalent to applying SYMMLQ to the normal equations

(A'A + λ² I) x = A'b

but is more stable.

Main features

the solution estimate is updated along orthogonal directions

the norm of the solution estimate ‖xᴸₖ‖₂ is increasing

the error ‖eₖ‖₂ := ‖xᴸₖ - x*‖₂ is decreasing

it is possible to transition cheaply from the LSLQ iterate to the LSQR iterate if there is an advantage (there always is in terms of error)

if A is rank deficient, identify the minimum least-squares solution

Input arguments

A::AbstractLinearOperator

b::Vector{Float64}

Optional arguments

M::AbstractLinearOperator=opEye(size(A,1)): a symmetric and positive definite dual preconditioner

N::AbstractLinearOperator=opEye(size(A,2)): a symmetric and positive definite primal preconditioner

sqd::Bool=false indicates whether or not we are solving a symmetric and quasi-definite augmented system If sqd = true, we solve the symmetric and quasi-definite system

[ E   A' ] [ r ]   [ b ]
[ A  -F  ] [ x ] = [ 0 ],

where E = M⁻¹ and F = N⁻¹.

If sqd = false, we solve the symmetric and indefinite system

[ E   A' ] [ r ]   [ b ]
[ A   0  ] [ x ] = [ 0 ].

In this case, N can still be specified and indicates the norm in which x and the forward error should be measured.

λ::Float64=0.0 is a regularization parameter (see the problem statement above)

σ::Float64=0.0 is an underestimate of the smallest nonzero singular value of A–-setting σ too large will result in an error in the course of the iterations

atol::Float64=1.0e-8 is a stopping tolerance based on the residual

btol::Float64=1.0e-8 is a stopping tolerance used to detect zero-residual problems

etol::Float64=1.0e-8 is a stopping tolerance based on the lower bound on the error

window::Int=5 is the number of iterations used to accumulate a lower bound on the error

utol::Float64=1.0e-8 is a stopping tolerance based on the upper bound on the error

itmax::Int=0 is the maximum number of iterations (0 means no imposed limit)

conlim::Float64=1.0e+8 is the limit on the estimated condition number of A beyond which the solution will be abandoned

verbose::Bool=false determines verbosity.

Return values

lslq() returns the tuple (x_lq, x_cg, err_lbnds, err_ubnds_lq, err_ubnds_cg, stats) where

x_lq::Vector{Float64} is the LQ solution estimate

x_cg::Vector{Float64} is the CG solution estimate (i.e., the LSQR point)

err_lbnds::Vector{Float64} is a vector of lower bounds on the LQ error–-the vector is empty if window is set to zero

err_ubnds_lq::Vector{Float64} is a vector of upper bounds on the LQ error–-the vector is empty if σ == 0 is left at zero

err_ubnds_cg::Vector{Float64} is a vector of upper bounds on the CG error–-the vector is empty if σ == 0 is left at zero

stats::SimpleStats collects other statistics on the run.

Stopping conditions

The iterations stop as soon as one of the following conditions holds true:

the optimality residual is sufficiently small (stats.status = "found approximate minimum least-squares solution") in the sense that either

‖Aᵀr‖ / (‖A‖ ‖r‖) ≤ atol, or

1 + ‖Aᵀr‖ / (‖A‖ ‖r‖) ≤ 1

an approximate zero-residual solution has been found (stats.status = "found approximate zero-residual solution") in the sense that either

‖r‖ / ‖b‖ ≤ btol + atol ‖A‖ * ‖xᴸ‖ / ‖b‖, or

1 + ‖r‖ / ‖b‖ ≤ 1

the estimated condition number of A is too large in the sense that either

1/cond(A) ≤ 1/conlim (stats.status = "condition number exceeds tolerance"), or

1 + 1/cond(A) ≤ 1 (stats.status = "condition number seems too large for this machine")

the lower bound on the LQ forward error is less than etol * ‖xᴸ‖

the upper bound on the CG forward error is less than utol * ‖xᶜ‖

References

R. Estrin, D. Orban and M. A. Saunders, Estimates of the 2-Norm Forward Error for SYMMLQ and CG, Cahier du GERAD G-2016-70, GERAD, Montreal, 2016. DOI http://dx.doi.org/10.13140/RG.2.2.19581.77288.

R. Estrin, D. Orban and M. A. Saunders, LSLQ: An Iterative Method for Linear Least-Squares with an Error Minimization Property, Cahier du GERAD G-2017-xx, GERAD, Montreal, 2017.

Krylov.lsmr — Function.

Solve the regularized linear least-squares problem

minimize ‖b - Ax‖₂² + λ² ‖x‖₂²

using the LSMR method, where λ ≥ 0 is a regularization parameter. LSQR is formally equivalent to applying MINRES to the normal equations

(A'A + λ² I) x = A'b

(and therefore to CRLS) but is more stable.

LSMR produces monotonic residuals ‖r‖₂ and optimality residuals ‖A'r‖₂. rt is formally equivalent to CRLS, though can be slightly more accurate.

Preconditioners M and N may be provided in the form of linear operators and are assumed to be symmetric and positive definite. If sqd is set to true, we solve the symmetric and quasi-definite system

[ E A' ] [ r ] [ b ] [ A -F ] [ x ] = [ 0 ],

where E = M⁻¹ and F = N⁻¹.

If sqd is set to false (the default), we solve the symmetric and indefinite system

[ E A' ] [ r ] [ b ] [ A 0 ] [ x ] = [ 0 ].

In this case, N can still be specified and indicates the norm in which x should be measured.

Krylov.lsqr — Function.

Solve the regularized linear least-squares problem

minimize ‖b - Ax‖₂² + λ² ‖x‖₂²

using the LSQR method, where λ ≥ 0 is a regularization parameter. LSQR is formally equivalent to applying CG to the normal equations

(A'A + λ² I) x = A'b

(and therefore to CGLS) but is more stable.

LSQR produces monotonic residuals ‖r‖₂ but not optimality residuals ‖A'r‖₂. It is formally equivalent to CGLS, though can be slightly more accurate.

Preconditioners M and N may be provided in the form of linear operators and are assumed to be symmetric and positive definite. If sqd is set to true, we solve the symmetric and quasi-definite system

[ E A' ] [ r ] [ b ] [ A -F ] [ x ] = [ 0 ],

where E = M⁻¹ and F = N⁻¹.

If sqd is set to false (the default), we solve the symmetric and indefinite system

[ E A' ] [ r ] [ b ] [ A 0 ] [ x ] = [ 0 ].

In this case, N can still be specified and indicates the norm in which x should be measured.

Krylov.minres — Function.

Solve the shifted linear least-squares problem

minimize ‖b - (A + λ I)x‖₂²

or the shifted linear system

(A + λ I) x = b

using the MINRES method, where λ ≥ 0 is a shift parameter, where A is square and symmetric.

MINRES is formally equivalent to applying CR to Ax=b when A is positive definite, but is typically more stable and also applies to the case where A is indefinite.

MINRES produces monotonic residuals ‖r‖₂ and optimality residuals ‖A'r‖₂.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite.

Krylov.preallocated_LinearOperator

Krylov.preallocated_symmetric_LinearOperator

Krylov.roots_quadratic — Function.

Find the real roots of the quadratic

q(x) = q₂ x² + q₁ x + q₀,

where q₂, q₁ and q₀ are real. Care is taken to avoid numerical cancellation. Optionally, nitref steps of iterative refinement may be performed to improve accuracy. By default, nitref=1.

Krylov.sym_givens — Function.

Numerically stable symmetric Givens reflection. Given a and b, return (c, s, ρ) such that

[ c  s ] [ a ] = [ ρ ]
[ s -c ] [ b ] = [ 0 ].

Krylov.symmlq — Function.

Solve the shifted linear system

(A + λ I) x = b

using the SYMMLQ method, where λ is a shift parameter, and A is square and symmetric.

SYMMLQ produces monotonic errors ‖x*-x‖₂.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite.

Krylov.to_boundary — Function.

Given a trust-region radius radius, a vector x lying inside the trust-region and a direction d, return σ1 and σ2 such that

‖x + σi d‖ = radius, i = 1, 2

in the Euclidean norm. If known, ‖x‖² may be supplied in xNorm2.

If flip is set to true, σ1 and σ2 are computed such that

‖x - σi d‖ = radius, i = 1, 2.

Krylov.vec2str