Docstrings

Reinforce.AbstractBatchIterator

Reinforce.AbstractBatchView

Reinforce.AbstractDataIterator

Reinforce.AbstractEnvironment

Reinforce.AbstractHistogram

Reinforce.AbstractObsIterator

Reinforce.AbstractObsView

Reinforce.AbstractPolicy

Reinforce.AbstractSet

Reinforce.AbstractState

Reinforce.AbstractWeights

Reinforce.AnalyticWeights

Reinforce.BatchIterator

Reinforce.CartPole

Reinforce.CartPoleV0

Reinforce.CartPoleV1

Reinforce.CoefTable

Reinforce.ConvergenceException

Reinforce.Cost

Reinforce.CrossEntropyMethod

Reinforce.DataIterator

Reinforce.DataView

Reinforce.DiscreteSet

Reinforce.DistanceLoss

Reinforce.ECDF

Reinforce.Episode

Reinforce.Episodes

Reinforce.EpisodicActorCritic

Reinforce.FrequencyWeights

Reinforce.Histogram

Reinforce.History

Reinforce.IntervalSet

Reinforce.KeyboardAction

Reinforce.KeyboardActionSet

Reinforce.L1dist

Reinforce.L2dist

Reinforce.LearnBase

Reinforce.Learnable

Reinforce.Linfdist

Reinforce.Loss

Reinforce.MarginLoss

Reinforce.Minimizable

Reinforce.MountainCarEnv

Reinforce.MouseAction

Reinforce.MouseActionSet

Reinforce.MultiArmedBanditEnv.MultiArmedBandit — Type.

The multi-armed bandit environment.

MultiArmedBandit(k, n = 100; σ = 1)

k is the number of available arms. The reward distribution for each arms is a Normal distribution centered at the range of [-1, 1] and the standard deviation is 1.

n is the max steps for each episode.

σ controls the the standard deviation for all Normal distribution.

MultiArmedBandit(x::Vector{<:Distribution}...)

In case that you want to other distributions as the reward distribution.

Reinforce.MultiArmedBanditEnv

Reinforce.ObsDim

Reinforce.ObsIterator

Reinforce.OnlineActorCritic

Reinforce.OnlineGAE

Reinforce.Penalty

Reinforce.Pendulum

Reinforce.ProbabilityWeights

Reinforce.RandomPolicy

Reinforce.RegressionModel

Reinforce.Reinforce

Reinforce.StateVector — Type.

A StateVector holds both the functions which will populate the state, and the most recent state.

Reinforce.StatisticalModel

Reinforce.StatsBase

Reinforce.StochasticTransformation

Reinforce.SupervisedLoss

Reinforce.Transformation

Reinforce.TupleSet

Reinforce.UnsupervisedLoss

Reinforce.Weights

Reinforce._done

Reinforce._start!

Reinforce.action — Function.

a = action(policy, r, s, A)

Take in the last reward r, current state s, and set of valid actions A = actions(env, s), then return the next action a.

Note that a policy could do a 'sarsa-style' update simply by saving the last state and action (s,a).

Reinforce.actions — Function.

A = actions(env, s)

Return a list/set/description of valid actions from state s.

Reinforce.addcounts!

Reinforce.addgrad!

Reinforce.adjr2

Reinforce.adjr²

Reinforce.aic

Reinforce.aicc

Reinforce.angle_normalize

Reinforce.autocor

Reinforce.autocor!

Reinforce.autocov

Reinforce.autocov!

Reinforce.aweights

Reinforce.bic

Reinforce.coef

Reinforce.coefnames

Reinforce.coeftable

Reinforce.competerank

Reinforce.confint

Reinforce.cor2cov

Reinforce.corkendall

Reinforce.corspearman

Reinforce.counteq

Reinforce.countmap

Reinforce.countne

Reinforce.counts

Reinforce.cov2cor

Reinforce.crosscor

Reinforce.crosscor!

Reinforce.crosscov

Reinforce.crosscov!

Reinforce.crossentropy

Reinforce.datasubset

Reinforce.denserank

Reinforce.deriv

Reinforce.deriv!

Reinforce.deriv2

Reinforce.describe

Reinforce.deviance

Reinforce.dof

Reinforce.dof_residual

Reinforce.ecdf

Reinforce.entropy

Reinforce.eval

Reinforce.findat

Reinforce.finished

Reinforce.fit

Reinforce.fit!

Reinforce.fitted

Reinforce.force_mag

Reinforce.fweights

Reinforce.genmean

Reinforce.genvar

Reinforce.geomean

Reinforce.getobs

Reinforce.getobs!

Reinforce.gkldiv

Reinforce.grad

Reinforce.grad!

Reinforce.gravity

Reinforce.harmmean

Reinforce.hist

Reinforce.include

Reinforce.indexmap

Reinforce.indicatormat

Reinforce.informationmatrix

Reinforce.inputdomain

Reinforce.inverse_rle

Reinforce.iqr

Reinforce.isclasscalibrated

Reinforce.isclipable

Reinforce.isconvex

Reinforce.isdifferentiable

Reinforce.isdistancebased

Reinforce.isfishercons

Reinforce.isfitted

Reinforce.islinear

Reinforce.islipschitzcont

Reinforce.islipschitzcont_deriv

Reinforce.islocallylipschitzcont

Reinforce.ismarginbased

Reinforce.ismdp — Function.

ismdp(env)::Bool

An environment may be fully observable (MDP) or partially observable (POMDP). In the case of a partially observable environment, the state s is really an observation o. To maintain consistency, we call everything a state, and assume that an environment is free to maintain additional (unobserved) internal state.

The ismdp query returns true when the environment is MDP, and false otherwise.

Reinforce.isminimizable

Reinforce.isnemitski

Reinforce.isstrictlyconvex

Reinforce.isstronglyconvex

Reinforce.issymmetric

Reinforce.istwicedifferentiable

Reinforce.isunivfishercons

Reinforce.kldivergence

Reinforce.kurtosis

Reinforce.learn

Reinforce.learn!

Reinforce.length_state

Reinforce.levelsmap

Reinforce.leverage

Reinforce.loglikelihood

Reinforce.mad

Reinforce.mad!

Reinforce.mass_cart

Reinforce.mass_pole

Reinforce.mass_pole_length

Reinforce.max_speed

Reinforce.max_torque

Reinforce.maxad

Reinforce.maxsteps — Function.

maxsteps(env)::Int

Return the max steps in single episode. Default is 0 (unlimited).

Reinforce.mean_and_cov

Reinforce.mean_and_std

Reinforce.mean_and_var

Reinforce.meanad

Reinforce.meanderiv

Reinforce.meanresponse

Reinforce.meanvalue

Reinforce.middle

Reinforce.midpoints

Reinforce.mode

Reinforce.model_response

Reinforce.modelmatrix

Reinforce.modes

Reinforce.moment

Reinforce.msd

Reinforce.mss

Reinforce.nobs

Reinforce.norepeats

Reinforce.nquantile

Reinforce.nulldeviance

Reinforce.nullloglikelihood

Reinforce.obs_dim

Reinforce.ordinalrank

Reinforce.pacf

Reinforce.pacf!

Reinforce.params

Reinforce.params!

Reinforce.partialcor

Reinforce.percentile

Reinforce.pole_length

Reinforce.predict

Reinforce.predict!

Reinforce.proportionmap

Reinforce.proportions

Reinforce.prox

Reinforce.prox!

Reinforce.psnr

Reinforce.pweights

Reinforce.r2

Reinforce.renyientropy

Reinforce.reset! — Function.

reset!(env) -> env

Reset an environment.

Reinforce.residuals

Reinforce.response

Reinforce.reward — Function.

r = reward(env)

Return the current reward of the environment.

Reinforce.rle

Reinforce.rmsd

Reinforce.rss

Reinforce.run_episode — Function.

run_episode(f, env, policy)
run_episode(env, policy) do (s, a, r, s′)
  # render or something else
end

Helper function for running an episode, and return the total reward gained in this episode.

Reinforce.r²

Reinforce.sample

Reinforce.sample!

Reinforce.samplepair

Reinforce.scaled

Reinforce.scattermat

Reinforce.score

Reinforce.sem

Reinforce.skewness

Reinforce.span

Reinforce.sqL2dist

Reinforce.state — Function.

s = state(env)

Return the current state of the environment.

retreive the most recently calculated state

Reinforce.state! — Function.

update the state, then return it

Reinforce.stderror

Reinforce.step! — Function.

r, s′ = step!(env, s, a)

Move the simulation forward, collecting a reward and getting the next state.

Reinforce.sumderiv

Reinforce.summarystats

Reinforce.sumvalue

Reinforce.targetdomain

Reinforce.targets

Reinforce.tiedrank

Reinforce.total_mass

Reinforce.totalvar

Reinforce.transform

Reinforce.transform!

Reinforce.trim

Reinforce.trim!

Reinforce.trimvar

Reinforce.update

Reinforce.update!

Reinforce.value

Reinforce.value!

Reinforce.value_deriv

Reinforce.value_deriv!

Reinforce.value_grad

Reinforce.value_grad!

Reinforce.variation

Reinforce.vcov

Reinforce.weights

Reinforce.winsor

Reinforce.winsor!

Reinforce.wmean

Reinforce.wmean!

Reinforce.wmedian

Reinforce.wquantile

Reinforce.wsample

Reinforce.wsample!

Reinforce.wsum

Reinforce.wsum!

Reinforce.x_threshold

Reinforce.zscore

Reinforce.zscore!

Reinforce.θ_threshold

Reinforce.τ