Docstrings

UnitfulAtomic.Eh_au

UnitfulAtomic.Ry

UnitfulAtomic.UnitfulAtomic

UnitfulAtomic.__init__

UnitfulAtomic.a0_au

UnitfulAtomic.auconvert — Function.

auconvert(x::Unitful.Quantity)

Convert a quantity to the appropriate atomic unit.

Examples

julia> auconvert(13.6u"eV")
0.4997907858599377 Eₕ

julia> auconvert(20u"nm")
377.94522509156565 a₀

auconvert(u::Unitful.Units, x::Number)

Interpret x as a quantity given in atomic units and convert it to the unit u.

Examples

julia> auconvert(u"eV", 1)  # convert 1 Eₕ to eV
27.211386013449417 eV

julia> auconvert(u"m", 1)   # convert 1 a₀ to m
5.2917721067e-11 m

UnitfulAtomic.aunit — Function.

aunit(x::Unitful.Quantity)
aunit(x::Unitful.Units)
aunit(x::Unitful.Dimensions)

Returns the appropriate atomic unit (a Unitful.Units object) for the dimension of x.

Examples

julia> aunit(2.3u"cm")
a₀

julia> aunit(u"T")
a₀^-2 e^-1 ħ

UnitfulAtomic.austrip — Function.

austrip(x::Unitful.Quantity)

Returns the value of the quantity converted to atomic units as a number type (i.e., with the units removed). This is equivalent to Unitful.ustrip(auconvert(x)).

Examples

julia> austrip(13.6u"eV")
0.4997907858599377

julia> austrip(20u"nm")
377.94522509156565

UnitfulAtomic.bohr

UnitfulAtomic.e_au

UnitfulAtomic.eval

UnitfulAtomic.hartree

UnitfulAtomic.include

UnitfulAtomic.k_au

UnitfulAtomic.localunits

UnitfulAtomic.me_au

UnitfulAtomic.ħ_au

UnitfulAtomic.μ_N