Docstrings

DandelionWebSockets.@expect — Macro.

Expect a call with some arguments, and perform an action when that call has been made..

DandelionWebSockets.@mock — Macro.

Define a mock type, given a symbol for its name and a type from which it should inherit.

DandelionWebSockets.@mockfunction — Macro.

Define one or more functions that mock the behaviour the object should have.

The functions defined here are the functions called by the code under test. The functions check that their arguments match those that are expected, and that the mock action is performed.

DandelionWebSockets.AbnormalNoCloseResponseReceived — Type.

No Close frame response was received by the client after a reasonable time

DandelionWebSockets.AbnormalSocketNotClosedByServer — Type.

The socket did not close the socket in a reasonable timeframe.

DandelionWebSockets.AbstractBackoff

DandelionWebSockets.AbstractClientProtocol

DandelionWebSockets.AbstractFrameWriter

DandelionWebSockets.AbstractHandshakeResult — Type.

AbstractHandshakeResult is a supertype for either a successful or an unsuccessful handshake.

DandelionWebSockets.AbstractMatcher — Type.

Abstract type for objects that match actual arguments against expected arguments. For instance, we sometimes don't care what value is provided a mock function, only what type it is. Also, this can be used to parse an actual argument as JSON and compare the resulting object, rather than relying on a string comparison.

All AbstractMatcher types T must define a function mock_match(::T, v::Any).

DandelionWebSockets.AbstractPinger

DandelionWebSockets.AbstractPonger

DandelionWebSockets.AbstractRetry

DandelionWebSockets.AbstractServerReader

DandelionWebSockets.AbstractWSClient

DandelionWebSockets.Backoff — Type.

A backoff that follows a atan curve, and reaches about 90% of max backoff in 12 attempts.

DandelionWebSockets.BadHandshake — Type.

BadHandshake is returned when an unsuccessful handshake has been made.

DandelionWebSockets.BinaryFrameSender — Type.

FrameSender is used to send multi-frame messages.

You are provided a FrameSender via the client. You then send data using the sendframe(sender, data; isfinal=false) method. For the last frame you set isfinal = true.

After the last frame has been sent, you MAY NOT use the FrameSender again. Trying to send another frame with the same instance of FrameSender WILL lead to an exception of type FinalFrameAlreadySentException.

NOTE: While you have a FrameSender in which you have not sent the last frame, then you MAY NOT use the client to send any other messages. Interleaving messages is prohibited by the WebSocket protocol and will lead to the socket being closed.

NOTE: You may send frames that are individually invalid UTF-8. However, the complete message, which is all frames concatenated, MUST be valid UTF-8, or the other endpoint is required to fail the WebSocket connection. With the TextFrameSender you can use both a String or a Vector{UInt8} as the payload. The Vector{UInt8} alternative can be used to send invalid UTF-8, or possibly invalid UTF-8.

Example

sender = sendmultiframetext(client)
sendframe(sender, "Hello")
sendframe(sender, "world")
sendframe(sender, "Goodbye"; isfinal=true)

Example

sender = sendmultiframebinary(client)
sendframe(sender, b"Hello")
sendframe(sender, b"world")
sendframe(sender, b"Goodbye"; isfinal=true)

DandelionWebSockets.CLOSE_STATUS_ABNORMAL_CLOSE

DandelionWebSockets.CLOSE_STATUS_EXPECTED_EXTENSION

DandelionWebSockets.CLOSE_STATUS_FATAL_UNEXPECTED_CONDITION

DandelionWebSockets.CLOSE_STATUS_GOING_AWAY

DandelionWebSockets.CLOSE_STATUS_INCONSISTENT_DATA

DandelionWebSockets.CLOSE_STATUS_MESSAGE_TOO_BIG

DandelionWebSockets.CLOSE_STATUS_NORMAL

DandelionWebSockets.CLOSE_STATUS_NO_STATUS

DandelionWebSockets.CLOSE_STATUS_POLICY_VIOLATION

DandelionWebSockets.CLOSE_STATUS_PROTOCOL_ERROR

DandelionWebSockets.CLOSE_STATUS_RESERVED_1004

DandelionWebSockets.CLOSE_STATUS_TLS_HANDSHAKE_FAILURE

DandelionWebSockets.CLOSE_STATUS_UNACCEPTABLE_DATA

DandelionWebSockets.ClientInitiatedCloseBehaviour — Type.

Closing the WebSocket connection is a procedure for closing the connection during the normal course the protocol lifetime.

DandelionWebSockets.ClientPingRequest — Type.

Send a ping request to the server.

DandelionWebSockets.ClientProtocol — Type.

Type for the logic of a client WebSocket.

DandelionWebSockets.ClientProtocolInput — Type.

Abstract type for all commands sent to ClientProtocol.

These commands are sent as arguments to the different handle functions on ClientProtocol. Each command represents an action on a WebSocket, such as sending a text frame, ping request, or closing the connection.

DandelionWebSockets.CloseRequest — Type.

A request to close the WebSocket.

DandelionWebSockets.CloseStatus — Type.

CloseStatusCode indicates a reason for closing the connection. It is optionally sent as the first two bytes of a Close frames payload.

DandelionWebSockets.CloseStatusAndReason

DandelionWebSockets.ClosingBehaviour

DandelionWebSockets.DandelionWebSockets

DandelionWebSockets.FailTheConnectionBehaviour — Type.

Failing the WebSocket connection is an action taken at certain points in the protocol specification, in response to error conditions.

The closing behaviour is to optionally send a Close frame, with an appropriate status code, and then close the socket.

DandelionWebSockets.FinalFrameAlreadySentException — Type.

FinalFrameAlreadySentException is thrown when sendframe is called again after the last frame.

DandelionWebSockets.Frame

DandelionWebSockets.FrameFromServer — Type.

A frame was received from the server.

DandelionWebSockets.FrameSender — Type.

FrameSender is used to send multi-frame messages.

You are provided a FrameSender via the client. You then send data using the sendframe(sender, data; isfinal=false) method. For the last frame you set isfinal = true.

After the last frame has been sent, you MAY NOT use the FrameSender again. Trying to send another frame with the same instance of FrameSender WILL lead to an exception of type FinalFrameAlreadySentException.

NOTE: While you have a FrameSender in which you have not sent the last frame, then you MAY NOT use the client to send any other messages. Interleaving messages is prohibited by the WebSocket protocol and will lead to the socket being closed.

NOTE: You may send frames that are individually invalid UTF-8. However, the complete message, which is all frames concatenated, MUST be valid UTF-8, or the other endpoint is required to fail the WebSocket connection. With the TextFrameSender you can use both a String or a Vector{UInt8} as the payload. The Vector{UInt8} alternative can be used to send invalid UTF-8, or possibly invalid UTF-8.

Example

sender = sendmultiframetext(client)
sendframe(sender, "Hello")
sendframe(sender, "world")
sendframe(sender, "Goodbye"; isfinal=true)

Example

sender = sendmultiframebinary(client)
sendframe(sender, b"Hello")
sendframe(sender, b"world")
sendframe(sender, b"Goodbye"; isfinal=true)

DandelionWebSockets.FrameWriter — Type.

FrameWriter is used by the protocols to write frames to a socket.

This is separate from the client protocol code, because the protocol for closing a connection is separate from the rest of the client protocol, and both need to send frames.

DandelionWebSockets.GoodHandshake — Type.

GoodHandshake is returned when a successful handshake has been made.

DandelionWebSockets.HTTPAdapter — Type.

HTTPAdapter is an abstract type for an HTTP package that can be used for a handshake GET.

Any subtype of HTTPAdapter should implement the dohandshake method.

DandelionWebSockets.HTTPHandshake — Type.

HTTPHandshake implements the HTTP handshake from the WebSocket specification.

It requires a random number generator to generate the WebSocket random key used to verify that the other side is actually a WebSocket server. The HTTPAdapter supplied implements the actual HTTP GET request, which returns an HTTPUpgradeResponse.

DandelionWebSockets.HTTPHandshakeLogic

DandelionWebSockets.HTTPUpgradeResponse — Type.

HTTPUpgradeResponse is a collection of data returned by a HTTP upgrade, regardless of package.

DandelionWebSockets.HTTPjlAdapter

DandelionWebSockets.HandshakeValidationResult

DandelionWebSockets.HeaderList

DandelionWebSockets.MockCall — Type.

An expected function call, along with expected arguments, and an action it should perform.

DandelionWebSockets.MockExpectationException — Type.

An exception thrown when an unexpected argument was found or a function was called.

DandelionWebSockets.OPCODE_BINARY

DandelionWebSockets.OPCODE_CLOSE

DandelionWebSockets.OPCODE_CONTINUATION

DandelionWebSockets.OPCODE_PING

DandelionWebSockets.OPCODE_PONG

DandelionWebSockets.OPCODE_TEXT

DandelionWebSockets.Opcode

DandelionWebSockets.Pinger

DandelionWebSockets.PongMissed — Type.

A pong reply was expected, but never received.

DandelionWebSockets.Ponger

DandelionWebSockets.Proxy

DandelionWebSockets.RandomizedBackoff — Type.

Randomize another backoff by adding some randomness to the backoff time.

DandelionWebSockets.Retry — Type.

Start a timer for some function, based on a backoff.

DandelionWebSockets.STATE_CLOSED

DandelionWebSockets.STATE_CLOSING

DandelionWebSockets.STATE_CONNECTING

DandelionWebSockets.STATE_OPEN

DandelionWebSockets.SendBinaryFrame — Type.

Send a binary frame, sent to ClientProtocol.

DandelionWebSockets.SendTextFrame — Type.

Send a text frame, sent to ClientProtocol.

DandelionWebSockets.ServerInitiatedCloseBehaviour — Type.

The server can initiate a Close, in which case this behaviour ensures a proper close.

DandelionWebSockets.ServerReader — Type.

Reading from a network socket and placing the resulting frame on a channel.

DandelionWebSockets.SocketClosed — Type.

Used when the underlying network socket was closed.

DandelionWebSockets.SocketState — Type.

Enum value for the different states a WebSocket can be in.

DandelionWebSockets.StopTaskException — Type.

An exception thrown into a task in order to stop it.

DandelionWebSockets.SuccessfulHandshake

DandelionWebSockets.TLSBufferedIO — Type.

TLSBufferedIO adapts a TLS socket so we can do byte I/O.

The stream returned by MbedTLS when using a TLS socket does not support the byte I/O used when reading a frame. It only supports reading a chunk of data. This is a fake stream that buffers some data and lets us do byte I/O.

Note: This should have been done by the BufferedStreams.jl package. However, I couldn't get it to work with the MbedTLS stream, for reasons unknown. If we can investigate and fix that problem, then we should really replace this type with a BufferedInputStream.

DandelionWebSockets.TextFrameSender — Type.

FrameSender is used to send multi-frame messages.

You are provided a FrameSender via the client. You then send data using the sendframe(sender, data; isfinal=false) method. For the last frame you set isfinal = true.

After the last frame has been sent, you MAY NOT use the FrameSender again. Trying to send another frame with the same instance of FrameSender WILL lead to an exception of type FinalFrameAlreadySentException.

NOTE: While you have a FrameSender in which you have not sent the last frame, then you MAY NOT use the client to send any other messages. Interleaving messages is prohibited by the WebSocket protocol and will lead to the socket being closed.

NOTE: You may send frames that are individually invalid UTF-8. However, the complete message, which is all frames concatenated, MUST be valid UTF-8, or the other endpoint is required to fail the WebSocket connection. With the TextFrameSender you can use both a String or a Vector{UInt8} as the payload. The Vector{UInt8} alternative can be used to send invalid UTF-8, or possibly invalid UTF-8.

Example

sender = sendmultiframetext(client)
sendframe(sender, "Hello")
sendframe(sender, "world")
sendframe(sender, "Goodbye"; isfinal=true)

Example

sender = sendmultiframebinary(client)
sendframe(sender, b"Hello")
sendframe(sender, b"world")
sendframe(sender, b"Goodbye"; isfinal=true)

DandelionWebSockets.Throws — Type.

Tells a mock function that it should throw an exception.

DandelionWebSockets.TypeMatcher — Type.

Match a value by checking its type only.

DandelionWebSockets.UniformResourceIdentifiers

DandelionWebSockets.UnsuccessfulHandshake

DandelionWebSockets.ValueMatcher — Type.

Simply match a value by equality. This is the default.

DandelionWebSockets.WSClient — Type.

A WebSocket client, used to connect to a server, and send messages.

Note: The keyword arguments in the constructor are primarily for testing.

DandelionWebSockets.WebSocketConnection

DandelionWebSockets.WebSocketConnectionImpl

DandelionWebSockets.WebSocketHandler

DandelionWebSockets.WebSocketHandshake — Type.

WebSocketHandshake represents a way to make a WebSocket connection handshake.

The only handshake detailed in the specification is an HTTP handshake, represented by the HTTPHandshake type (regardless of HTTP package used to implement it). The specification does mention the possibility of other types of handshakes, even though these would be outside of the specification.

DandelionWebSockets.WebSocketUpgrade

DandelionWebSockets.WebSocketUpgradeAdapter

DandelionWebSockets._expectheader

DandelionWebSockets.attach

DandelionWebSockets.backoff_max

DandelionWebSockets.backoff_min

DandelionWebSockets.buffer

DandelionWebSockets.clientprotocolinput

DandelionWebSockets.closesocket

DandelionWebSockets.closestatusandreason

DandelionWebSockets.closetheconnection

DandelionWebSockets.connection_result_ — Function.

connection_result(::WSClient, ::GoodHandshake, ::WebSocketHandler, fix_small_message_latency::Bool)

For a valid handshake, start all background tasks for this connection. This includes tasks for reading and writing from the socket, as well as a task for user callback.

The HTTP Upgrade failed, for whatever reason.

DandelionWebSockets.default_timer_factory

DandelionWebSockets.do_reader — Function.

Read frames from the network, until an exception is thrown in this task.

DandelionWebSockets.dohandshake — Function.

dohandshake(http::HTTPAdapter, uri::String, headers::HeaderList)

Do an HTTP GET request to uri including headers headers.

The headers list will contain all WebSocket upgrade specific headers, such as Connection, Upgrade, and Sec-WebSocket-Key.

DandelionWebSockets.eval

DandelionWebSockets.failtheconnection

DandelionWebSockets.fill_buffer — Function.

Read all available data, and block until we have enough to fulfíll the next read.

DandelionWebSockets.getrequestheaders

DandelionWebSockets.handle — Function.

Send a single text frame.

Send a single binary frame.

Handle a user request to close the WebSocket.

The underlying socket was closed. This is sent by the reader.

Handle a frame from the server.

DandelionWebSockets.handle_binary

DandelionWebSockets.handle_close

DandelionWebSockets.handle_continuation

DandelionWebSockets.handle_ping

DandelionWebSockets.handle_pong

DandelionWebSockets.handle_text

DandelionWebSockets.include

DandelionWebSockets.isclosedcleanly

DandelionWebSockets.iscontrolframe

DandelionWebSockets.issuccessful — Function.

issuccessful(::AbstractHandshakeResult)

True if a handshake was succcessful, false otherwise.

DandelionWebSockets.masking!

DandelionWebSockets.maybewrapio

DandelionWebSockets.mock_action

DandelionWebSockets.mock_match

DandelionWebSockets.name_args_ — Function.

Define a name for an argument, if no name is provided.

DandelionWebSockets.on_binary — Function.

Handle a binary frame.

DandelionWebSockets.on_text — Function.

Handle a text frame.

DandelionWebSockets.performhandshake — Function.

performhandshake(h::HTTPHandshake, uri::String)

Do a handshake with the server at uri, with parameters supplied by h. Validate the handshake, and return a good or bad result, depending on the validation.

DandelionWebSockets.ping_sent

DandelionWebSockets.pong_received

DandelionWebSockets.protocolstate — Function.

The state of the connection.

DandelionWebSockets.readstatusandreason

Base.reset — Function.

Reset the backoff to its initial state.

DandelionWebSockets.retry

DandelionWebSockets.send — Function.

Send a frame to the other endpoint, using the supplied payload and opcode.

DandelionWebSockets.send_binary — Function.

Send a single binary frame.

DandelionWebSockets.send_text — Function.

Send a single text frame.

DandelionWebSockets.sendcloseframe

DandelionWebSockets.sendframe — Function.

Send a frame with a payload to the other endpoint. isfinal must be set to true for the last frame.

After the call with isfinal = true, then this method MAY NOT be called again. If it is, then a FinalFrameAlreadySentException will be thrown.

Specialization of sendframe for TextFrameSender, for sending Strings as data.

DandelionWebSockets.sendmultiframebinary

DandelionWebSockets.sendmultiframetext

DandelionWebSockets.set_function

DandelionWebSockets.show

DandelionWebSockets.start_buffer

DandelionWebSockets.start_reader

DandelionWebSockets.state_closed — Function.

The WebSocket was closed.

DandelionWebSockets.state_closing — Function.

The WebSocket is about to close.

DandelionWebSockets.state_connecting — Function.

The WebSocket is trying to connect.

DandelionWebSockets.state_open — Function.

The WebSocket is open and ready to send and receive messages.

DandelionWebSockets.stop — Function.

Close the WebSocket connection.

DandelionWebSockets.tcpnodelay — Function.

tcpnodelay(::IO)

Sets the TCP_NODELAY flag on a socket. This is a separate function only for testing purposes. It can be implemented with a more specific type if the flag makes no sense for another IO subtype.

DandelionWebSockets.to_matcher

DandelionWebSockets.validateresponse

DandelionWebSockets.wsconnect — Function.

Connect the client to a WebSocket server at uri, and use handler for the callbacks.

Arguments

fix_small_message_latency::Bool = false: Set the TCP_NODELAY flag to improve small message latency.

Fix small message latency

The TCP protocol can buffer small messages (1448 bytes and smaller). The reason is that this reduces the overhead when sending large amounts of small packets. However, it also means that latency can be much higher for small messages. This buffering can be disabled by setting a flag TCPNODELAY. By default, the WebSocket client will now set the TCPNODELAY flag.

If your application will send and receive primarily small messages (1448 bytes or smaller), and it is sensitive to latency, then leave fix_small_message_latency set to true (now the default). This sets the TCPNODELAY flag. If you are not concerned about latency, but concerned about throughput for many small messages, then you can set `fixsmallmessagelatency = false`. Then you may get higher throughput, at the expense of higher latency for small messages.