Docstrings

RegionTrees.AbstractRefinery

RegionTrees.BodyAndFaceCenters

RegionTrees.Cell

RegionTrees.HyperRectangle

RegionTrees.RegionTrees

RegionTrees.TwosArray — Type.

Represents a static array with N dimensions whose size is exactly 2 along each dimension. This makes templating on the number of dimensions easier than with a regular SArray.

RegionTrees.adaptivesampling!

RegionTrees.allcells

RegionTrees.allleaves

RegionTrees.body_and_face_centers

RegionTrees.center

RegionTrees.child_boundary

RegionTrees.child_indices

RegionTrees.children

RegionTrees.eval

RegionTrees.faces

RegionTrees.findleaf

RegionTrees.include

RegionTrees.initial_data

RegionTrees.isleaf

RegionTrees.map_children

RegionTrees.needs_refinement

RegionTrees.parent

RegionTrees.refine_data

RegionTrees.split!

RegionTrees.split!_impl

RegionTrees.vertices