Docstrings

MurmurHash3.MurmurHash3 — Module.

MurmurHash3 was written by Austin Appleby, and is placed in the public domain. The author hereby disclaims copyright to this source code.

This version was translated into Julia by Scott P. Jones It is licensed under the MIT license

Note - The x86 and x64 versions do not produce the same results, as the algorithms are optimized for their respective platforms. You can still compile and run any of them on any platform, but your performance with the non-native version will be less than optimal.

MurmurHash3.add_utf8

MurmurHash3.add_utf8_split

MurmurHash3.c1

MurmurHash3.c2

MurmurHash3.d1

MurmurHash3.d2

MurmurHash3.dn12

MurmurHash3.dn18

MurmurHash3.dn6

MurmurHash3.e1

MurmurHash3.e2

MurmurHash3.e3

MurmurHash3.e4

MurmurHash3.eval

MurmurHash3.fmix

MurmurHash3.get_utf8

MurmurHash3.get_utf8_2

MurmurHash3.get_utf8_3

MurmurHash3.get_utf8_4

MurmurHash3.include

MurmurHash3.mask_load

MurmurHash3.mergebytes

MurmurHash3.mhblock

MurmurHash3.mhbody

MurmurHash3.mhfin

MurmurHash3.mhtail1

MurmurHash3.mhtail2

MurmurHash3.mmhash128_4

MurmurHash3.mmhash128_8_a

MurmurHash3.mmhash128_8_c

MurmurHash3.mmhash128_8_u

MurmurHash3.mmhash128_a

MurmurHash3.mmhash128_c

MurmurHash3.mmhash128_u

MurmurHash3.mmhash32

MurmurHash3.msk6

MurmurHash3.rotl

MurmurHash3.rotl13

MurmurHash3.rotl15

MurmurHash3.rotl16

MurmurHash3.rotl17

MurmurHash3.rotl18

MurmurHash3.rotl19

MurmurHash3.rotl27

MurmurHash3.rotl31

MurmurHash3.rotl33

MurmurHash3.shift_mix

MurmurHash3.shift_n

MurmurHash3.shift_n_32

MurmurHash3.u32

MurmurHash3.u64

MurmurHash3.u8

MurmurHash3.up16

MurmurHash3.up24

MurmurHash3.up32

MurmurHash3.up40

MurmurHash3.up48

MurmurHash3.up56

MurmurHash3.up8

MurmurHash3.xor13

MurmurHash3.xor16

MurmurHash3.xor33