Docstrings

UnsafeArrays.@gc_preserve — Macro.

UnsafeArrays.@gc_preserve A B ... expr

Protect A, B, ... from garbage collection while executing expr. Contrary to GC.@preserve, expr is executed in a new scope.

Equivalent to

GC.@preserve A B ... let
    expr
end

On Julia versions that do not provide GC.@preserve, a fallback implementation is used.

UnsafeArrays.@uviews — Macro.

@uviews A B ... expr

Replace arrays A, B, ... by uview(A), uview(B), ... during execution of expr, while protecting the original arrays from garbage collection.

Equivalent to

GC.@preserve A B ... begin
    let A = uview(A), B = uview(B), ...
        expr
    end
end

The unsafe views must not be allowed to escape the scope of expr. The original arrays must not be resized/appended/etc. during the execution of expr.

UnsafeArrays.DenseIdx

UnsafeArrays.IdxUnitRange

UnsafeArrays.UnsafeArray — Type.

UnsafeArray{T,N} <: DenseArray{T,N}

An UnsafeArray is an bitstype wrapper around a memory pointer and a size tuple. It's an intended to be used as a short-lived, allocation-free alternative to an an Array returned by unsafe_wrap or a SubArray returned by view or . Use with caution!

Constructors:

UnsafeArray{T,N}(pointer::Ptr{T}, size::NTuple{N,Int}) where {T,N}
UnsafeArray(pointer::Ptr{T}, size::NTuple{N,Int}) where {T,N}

UnsafeArray requires isbitstype(T) == true.

Note: It's safe to construct an empty multidimensional UnsafeArray:

UnsafeArray(Ptr{Int}(0), (0,...))

Note: You must ensure that A is not garbage collected or reallocated via (e.g.) resize!, sizehint! etc. while U is in use! Use only in situations where you have full control over the life cycle of A and U.

deepcopy(::UnsafeArray) returns a standard Array.

UnsafeArrays.UnsafeArrays

UnsafeArrays._fast_sub2ind

UnsafeArrays._maybe_unsafe_uview

UnsafeArrays._noinline_nop

UnsafeArrays._require_one_based_indexing

UnsafeArrays._sub_axes

UnsafeArrays._sub_size

UnsafeArrays._unsafe_view_impl

UnsafeArrays.eval

UnsafeArrays.include

UnsafeArrays.unsafe_uview — Function.

UnsafeArray.unsafe_uview(A::AbstractArray, I::Vararg{Base.ViewIndex,N})
UnsafeArray.unsafe_uview(A::AbstractArray, i::Base.ViewIndex)
UnsafeArray.unsafe_uview(A::AbstractArray)

To support uview for custom array types, add methods to unsafe_uview instead of uview. Implementing UnsafeArray.unsafe_uview(A::CustomArrayType) will often be sufficient.

It may be necessary to provide a custom implementation of Base.deepcopy for types with a custom implementation of unsafe_uview, to ensure that the result of deepcopy does NOT contain unsafe views.

UnsafeArrays.uview — Function.

uview(A::AbstractArray, I...)

Unsafe equivalent of view. May return an UnsafeArray, a standard SubArray or A itself, depending on I... and the type of A.

As uview may return an UnsafeArray, A itself and it's contents must be protected from garbage collection (e.g. via GC.@preserve on Julia > v0.6) and memory reallocation while the view is in use.

Use uviews(f::Function, As::AbstractArray...) or @uviews A ... expr to use unsafe views of one or multiple arrays with automatically GC protection.

uviews(A, B, ...) do (A_u, B_u, ...)
    # Do something with the unsafe views A_u, B_u, ...
    # Code here must not resize/append/etc. A, B, ...
end

To provide support for uview for custom array types, add methods to function UnsafeArrays.unsafe_uview.

UnsafeArrays.uviews — Function.

uviews(f::Function, As::AbstractArray...)

Equivalent to f(map(uview, As)...). Automatically protects the array(s) As from garbage collection during execution of f.

Example:

uviews(A, B, ...) do (A_u, B_u, ...)
    # Do something with the unsafe views A_u, B_u, ...
    # Code here must not resize/append/etc. A, B, ...
end

In many cases, it may be preferable to use @uviews insted of views.