Docstrings

Hyperspecialize.@concretization — Macro.

@concretization(typetag)

Return the set of types corresponding to a type tag. A type tag is a module-qualified symbol mod.:Key where mod specifies a module and :Key is a symbol. If just the :Key is given, then the module is assumed to be the module in which the macro was expanded. If no concretization exists, return nothing.

A concretization can be set and modified with @concretize and @widen

Examples

julia> @concretization(BestInts)
nothing

julia> @concretize BestInts [Int32, Int64]
Set(Type[Int32, Int64])

julia> @concretization BestInts
Set(Type[Int32, Int64])

Hyperspecialize.@concretize — Macro.

@concretize(typetag, types)

Define the set of types corresponding to a type tag as types, where types is either a single type or any collection that may be passed to a set constructor. A type tag is a module-qualified symbol mod.:Key where mod specifies a module and :Key is a symbol. If just the :Key is given, then the module is assumed to be the module in which the macro was expanded.

Note that you may not concretize a type in another module.

Examples

julia> @concretize Main.BestInts [Int32, Int64]
Set(Type[Int32, Int64])

julia> @concretize BestFloats Float64
Set(Type[Float64])

julia> @concretize BestStrings (String,)
Set(Type[String])

julia> @concretization BestInts
Set(Type[Int32, Int64])

@concretize(typetag)

If no concretization exists, define the set of types corresponding to typetag as the concrete subtypes of whatever type shares the name of Key at load time. typetag is a type tag, or a module-qualified symbol mod.:Key where mod specifies a module and :Key is a symbol. If just the :Key is given, then the module is assumed to be the module in which the macro was expanded.

Note that you may not concretize a type tag in another module.

Examples

julia> @concretization(Main.Real)
nothing

julia> @concretize(Main.Real)
Set(Type[BigInt, Bool, UInt32, Float64, Float32, Int64, Int128, Float16, UInt128, UInt8, UInt16, BigFloat, Int8, UInt64, Int16, Int32])

julia> @concretize BestInts [Int32, Int64]
Set(Type[Int32, Int64])

julia> @concretization BestInts
Set(Type[Int32, Int64])

julia> @concretize NotDefinedHere
ERROR: cannot create default concretization from type tag Main.NotDefinedHere: not defined.

Hyperspecialize.@isdefined

Hyperspecialize.@replicable — Macro.

@replicable block

Replicate the code in block where each type tag referred to by @hyperspecialize(typetag) is replaced by an element in the concretization of typetag. block is replicated at global scope in the module where @replicable was expanded once for each combination of types in the concretization of each typetag. A type tag is a module-qualified symbol mod.:Key where mod specifies a module and :Key is a symbol. If just the :Key is given, then the module is assumed to be the module in which the macro was expanded. If no concretization exists for a type tag, create a default concretization consisting of the conrete subtypes of whatever type shares the name of Key at load time.

If @widen is called for a type tag which has been referenced by a @replicable code block, then that code block will be replicated even more to reflect the new concretization.

Examples

julia> @concretize BestInts [Int32, Int64]
Set(Type[Int32, Int64])

julia> @replicable println(@hyperspecialize(BestInts), @hyperspecialize(BestInts))
Int32Int32
Int32Int64
Int64Int32
Int64Int64

julia> @widen BestInts (Bool,)
BoolBool
BoolInt32
BoolInt64
Int32Bool
Int64Bool
Set(Type[Bool, Int32, Int64])

Hyperspecialize.@widen — Macro.

@widen(typetag, types)

Expand the set of types corresponding to a type tag to include types, where types is either a single type or any collection that may be passed to a set constructor. A type tag is a module-qualified symbol mod.:Key where mod specifies a module and :Key is a symbol. If just the :Key is given, then the module is assumed to be the module in which the macro was expanded. If no concretization exists, create a default concretization consisting of the conrete subtypes of whatever type shares the name of Key at load time.

If @widen is called for a type tag which has been referenced by a @replicable code block, then that code block will be replicated even more to reflect the new concretization.

Examples

julia> @concretize BestInts [Int32, Int64]
Set(Type[Int32, Int64])

julia> @replicable println(@hyperspecialize(BestInts))
Int32
Int64

julia> @widen BestInts (Bool, Int32, UInt128)
Bool
UInt128
Set(Type[Bool, UInt128, Int32, Int64])

julia> @concretization BestInts
Set(Type[Bool, Int8, Int32, Int64, UInt128])

Hyperspecialize.Concretization

Hyperspecialize.Hyperspecialize

Hyperspecialize.Replicable

Hyperspecialize.State

Hyperspecialize._concretization

Hyperspecialize._concretize

Hyperspecialize._define

Hyperspecialize._get_hyperspecialize

Hyperspecialize._is_hyperspecialize

Hyperspecialize._replicable

Hyperspecialize._widen

Hyperspecialize.allsubtypes — Function.

allsubtypes(t)

Return an Array containing all subtypes of t at load time.

Examples

julia> Hyperspecialize.allsubtypes(Real)
24-element Array{Type,1}:
 Real
 AbstractFloat
 BigFloat
 Float16
 Float32
 Float64
 AbstractIrrational
 Irrational
 Integer
 Bool
 Signed
 BigInt
 Int128
 Int16
 Int32
 Int64
 Int8
 Unsigned
 UInt128
 UInt16
 UInt32
 UInt64
 UInt8
 Rational

Hyperspecialize.concretesubtypes — Function.

concretesubtypes(t)

Return an Array containing all concrete subtypes of t at load time.

Examples

julia> Hyperspecialize.concretesubtypes(Real)
16-element Array{Any,1}:
 BigFloat
 Float16
 Float32
 Float64
 Bool
 BigInt
 Int128
 Int16
 Int32
 Int64
 Int8
 UInt128
 UInt16
 UInt32
 UInt64
 UInt8

Hyperspecialize.eval

Hyperspecialize.include

Hyperspecialize.parse_element

Hyperspecialize.postwalk