Docstrings

SweepOperator.AMat

SweepOperator.AVec

SweepOperator.SweepOperator

SweepOperator.eval

SweepOperator.include

SweepOperator.sweep! — Function.

Symmetric sweep operator

Symmetric sweep operator of the matrix A on element k. A is overwritten. inv = true will perform the inverse sweep. Only the upper triangle is read and swept.

sweep!(A, k, inv = false)

Providing a Range, rather than an Integer, sweeps on each element in the range.

sweep!(A, first:last, inv = false)

Example:

x = randn(100, 10)
xtx = x'x
sweep!(xtx, 1)
sweep!(xtx, 1, true)

SweepOperator.sweep_with_buffer!