Docstrings

AsyPlots.Arrow — Type.

Arrow(name::AbstractString,
      size::Real,
      position::Real)

Store instructions for drawing an arrow

position is an element of [0,1] which indicates how far along the path the arrow should be drawn

NoArrow() returns a no-arrow instruction, while Arrow3() gives an arrow suitable for 3D paths

AsyPlots.Arrow3

AsyPlots.AsyPlots

AsyPlots.AsyString

AsyPlots.BoundingBox

AsyPlots.Circle

AsyPlots.Circle2D — Type.

Circle2D(center,radius; pen::Pen,fillpen::Pen)

A graphics primitive representing a circle in the plane

center may be a Vec2 or a 2-tuple of Reals or a Complex

Examples

julia> Circle2D((0,0),1;pen="LightBlue",fillpen="red")
Circle2D((0,0),1;pen=LightBlue,fillpen=red)

AsyPlots.CubicInterpolation

AsyPlots.GraphicElement

AsyPlots.GraphicElement2D

AsyPlots.GraphicElement3D

AsyPlots.GraphicString

AsyPlots.IT

AsyPlots.Juno

AsyPlots.Label

AsyPlots.Label2D

AsyPlots.Label3D

AsyPlots.LinearInterpolation

AsyPlots.NamedColor — Type.

NamedColor(name::String,color::RGB)

A named ColorTypes.RGB

Examples

julia> NamedColor("DarkGreen")
NamedColor("DarkGreen",(0.0, 0.39, 0.0))

AsyPlots.NoArrow

AsyPlots.NoPen

AsyPlots.Path — Function.

Path(A;kwargs)

Return a Path2D or Path3D object, as appropriate

Examples

julia> Path([1 2; 3 4])
Path2D(<2 points>)

AsyPlots.Path2D — Type.

Path2D(points;label="",pen=Pen(),arrow=NoArrow(),spline=false)
Path2D(x,y;label="",pen=Pen(),arrow=NoArrow(),spline=false)

A graphics primitive representing a two-dimensional path

points may be an Array of Vec2s, an Array of 2-tuples, or an $n × 2$ Array. Alternatively, iterables of coordinates may be supplied separately as x and y

Examples

julia> Path2D([(0,0),(1,0),(1,1)];pen="MidnightBlue")
Path2D(<3 points>;pen=MidnightBlue)

AsyPlots.Path3D — Type.

Path3D(points;label="",pen=Pen(),arrow="",spline=false)

A graphics primitive representing a two-dimensional path

points may be an Array of Vec3s or an Array of 3-tuples. Alternatively, iterables of coordinates may be supplied separately as x and y

Examples

julia> Path3D([(0,0),(1,0),(1,1)];pen="MidnightBlue")

AsyPlots.Pen — Type.

Pen(color::NamedColor,
    opacity::Real,
    linewidth::Real,
    other::String)

Store drawing properties

Examples

julia> Pen(color="DarkGreen",opacity=0.5)
Pen(color=DarkGreen,opacity=0.5)

AsyPlots.PixelMap

AsyPlots.Plot

AsyPlots.Plot2D — Type.

Plot2D(elements::Array{<:GraphicElement,1},
       options::Array{Any,1})

A container for a list of graphics primitives to draw, together with drawing options

Examples

julia> Plot([Path([0 0; 1 1]),Polygon([exp(2*pi*im*k/5) for k=1:5])])

AsyPlots.Plot3D — Type.

Plot3D(elements::Array{<:GraphicElement2D,1},
       options::Array{Any,1})

A container for a list of graphics primitives to draw, together with drawing options

Examples

julia> Plot(Path([0 0 0; 1 1 1]),Path([0 0 0; 0 0 1]))

AsyPlots.Point — Function.

Point(x::Real,y::Real;kwargs...)
Point(x::Real,y::Real,z::Real;kwargs...)
Point(P;kwargs...)

Return a Point2D or Point3D object, as appropriate

Examples

julia> Point(1,2;pen=Pen(opacity=0.5))
Point2D(1,2;pen=opacity(0.5))

AsyPlots.Point2D — Type.

Point2D(x::Real,y::Real; label="",pen=Pen())
Point2D(P; label="",pen=Pen())

A graphics primitive representing a two-dimensional point.

P may be a 2-tuple of real numbers, a Vec2, or a Complex

Examples

julia> Point2D(3,-1;pen="DarkGreen")
Point2D(3,-1;pen=DarkGreen)

AsyPlots.Point3D — Type.

Point3D(x::Real,y::Real,z::Real; label="",pen=Pen())
Point3D(P; label="",pen=Pen())

A graphics primitive representing a three-dimensional point.

P may be a 3-tuple of real numbers or a Vec3

Examples

julia> Point3D(0,4,5;pen="DarkGreen")

AsyPlots.Polygon — Function.

Polygon(points;kwargs...)

Return a Polygon2D or a Polygon3D as appropriate

points may be an array of Vec2s or Complexes, or an $n × 2$ array of Reals

Examples

julia> Polygon([im,0,1])
Polygon2D(<3 points>)

AsyPlots.Polygon2D — Type.

Polygon2D(points;pen=Pen(),
                 fillpen=Pen(color="white"),
                 spline=false)

A graphics primitive representing a two-dimensional polygon

points may be an Array of Vec2s, an Array of 2-tuples, or an $n × 2$ Array. Alternatively, iterables of coordinates may be supplied separately as x and y

Examples

julia> Polygon2D([(0,0),(1,0),(1,1)];pen="MidnightBlue")
Polygon2D(<3 points>;pen=MidnightBlue)

AsyPlots.Polygon3D — Type.

Polygon3D(points;pen=Pen(),
                 fillpen=Pen(color="white"),
                 spline=false)

A graphics primitive representing a three-dimensional polygon

points may be an Array of Vec3s or an Array of 3-tuples.

Examples

julia> Polygon3D([(0,0,0),(1,0,0),(1,1,0)];pen="MidnightBlue")

AsyPlots.RawString — Type.

RawString(s::AbstractString)

Container for directly inserting Asymptote drawing commands

Examples

julia> Plot([Circle((0,0),1),RawString2D("draw((0,0)--dir(20));")])

AsyPlots.RealOrComplex

AsyPlots.Scale

AsyPlots.Scale2D

AsyPlots.Shift

AsyPlots.Shift2D

AsyPlots.Surface — Type.

Surface(x::Array{<:Real},
        y::Array{<:Real},
        z::Array{<:Real,2};
        options)
Surface(z::Array{<:Real},2)

A graphics primitive representing a surface in three dimensions x and y may be one- or two-dimensional arrays

The surface passes through the points [x[i,j],y[i,j],z[i,j] for i=1:size(z,1),j=1:size(z,2)]

The options are

colors: A vector of color names, for coloring

spline: whether to draw a smooth or piecewise smooth surface

surfacepen: a pen for drawing the surface

meshpen: a pen for drawing the grid lines on the surface

clip: either false or a boolean array of the same dimensions as x, y, and z, specifying patches to exclude

AsyPlots.Transform

AsyPlots.Vec2

AsyPlots.Vec3

AsyPlots._BACKEND

AsyPlots._DEFAULT_ARROW_KWARGS

AsyPlots._DEFAULT_CIRCLE2D_KWARGS

AsyPlots._DEFAULT_HEATMAP_KWARGS

AsyPlots._DEFAULT_ISOLINE_KWARGS

AsyPlots._DEFAULT_LABEL2D_KWARGS

AsyPlots._DEFAULT_LABEL3D_KWARGS

AsyPlots._DEFAULT_PATH2D_KWARGS

AsyPlots._DEFAULT_PATH3D_KWARGS

AsyPlots._DEFAULT_PEN_ARGS

AsyPlots._DEFAULT_PIXELMAP_KWARGS

AsyPlots._DEFAULT_PLOT2D_KWARGS

AsyPlots._DEFAULT_PLOT3D_KWARGS

AsyPlots._DEFAULT_POINT2D_KWARGS

AsyPlots._DEFAULT_POINT3D_KWARGS

AsyPlots._DEFAULT_POLYGON2D_KWARGS

AsyPlots._DEFAULT_POLYGON3D_KWARGS

AsyPlots._DEFAULT_SURFACE_KWARGS

AsyPlots._DEFAULT_WIDTH

AsyPlots._FIT_DIMENSION

AsyPlots._HAVE_ASY

AsyPlots._JUNO_ASY_FORMAT

AsyPlots._JUNO_CAIRO_FORMAT

AsyPlots._SHOW_PLOTS

AsyPlots.__init__

AsyPlots._is_ijulia

AsyPlots._is_juno

AsyPlots.addtocontext!

AsyPlots.animate — Function.

animate(filename::AbstractString,
        plots::Vector{Plot2D};
        rate=10)

animate(plots::Vector{Plot2D};rate=10)

Make an .mp4 video from plots, with frame rate rate

If "filename" is given, the movie file will be stored there. Otherwise, the movie will be opened.

Example

X = cumsum(randn(100000))
Y = cumsum(randn(100000))
plots = [Plot(Path(X[1:t],Y[1:t])) for t=10:10:10000]
animate(plots)

AsyPlots.arrowhead

AsyPlots.aspectratio

AsyPlots.asy — Function.

asy()

Set the backend to Asymptote

AsyPlots.axes

AsyPlots.blue

AsyPlots.boundingbox

AsyPlots.box

AsyPlots.bytes

AsyPlots.cairo — Function.

cairo()

Set the backend to Cairo

AsyPlots.check_asy_present

AsyPlots.cmap

AsyPlots.colorplaces

AsyPlots.counterclockwise

AsyPlots.defaultcolor

AsyPlots.defaultwidth

AsyPlots.det3p

AsyPlots.distance — Function.

distance(x,y,a,b,c,d)

Return the distance from the point (x,y) to the line segment with endpoints (a,b) and (c,d)

AsyPlots.enclosequote — Function.

enclosequote(s)

Enclose s in quotation marks, unless it starts with "Label"

AsyPlots.eval

AsyPlots.extension

AsyPlots.filterjoin — Function.

filterjoin(args...)

Concatenate nonempty string representations of args, separated by commas

AsyPlots.fitheight — Function.

fitheight()

Set the figures appearing in the Juno plot pane to match its height

AsyPlots.fitsquare

AsyPlots.fitwidth — Function.

fitwidth()

Set the figures appearing in the Juno plot pane to match its width

AsyPlots.green

AsyPlots.heatmap

AsyPlots.histogram — Function.

histogram(data::Vector;
          bincount=30,
          bins=linspace(minimum(data),maximum(data),bincount),
          xlabel="",
          ylabel="",
          title="",
          pen=Pen(color=NamedColor("MidnightBlue"),opacity=0.3))

Make a histogram of data

AsyPlots.include

AsyPlots.interactive

AsyPlots.is_no_pen

AsyPlots.isclip

AsyPlots.isdefault

AsyPlots.isinside — Function.

isinside(p::Point,pointlist::AbstractArray{Point,1})

Determine whether p is inside pointlist.

This function was copied from Luxor.jl. It is an implementation of an algorithm due to Hormann and Agathos (2001)

AsyPlots.isolines

AsyPlots.iswellinside

AsyPlots.junodefault — Function.

junodefault()

Reset the default output formats for the Juno plot pane

AsyPlots.junopng — Function.

junopng()

Set the default output format for the Juno plot pane to png

AsyPlots.junosvg — Function.

junosvg()

Set the default output format for the Juno plot pane to svg

AsyPlots.kwargstring

AsyPlots.layout

AsyPlots.lwcorrect

AsyPlots.otherargs

AsyPlots.pairs

AsyPlots.piechart — Function.

piechart(labels,frequencies; title="")

labels: a vector of strings

frequencies: a vector of Reals

Return a piechart with each sector i labeled labels[i] and having central angle proportional to frequencies[i]

AsyPlots.plot — Function.

plot(x,y;kwargs...)
plot(y;kwargs...)

Return a graph of the path with $x$ and $y$ values given by x and y

x defaults to 0:length(y)-1. kwargs are applied to the Path2D object representing the line or to the containing Plot2D, as appropriate

plot(xs::Vector{<:Vector{<:Real}},
     ys::Vector{<:Vector{<:Real}};
     kwargs...)

Multiple line graphs in the same figure

plot(x,y,z;kwargs...)
plot(z::Array{<:Real,2};kwargs...)

A graph of the surface with $x$, $y$, and $z$ values x, y, and z

x defaults to [i-1 for i=1:size(z,1),j=1:size(z,2)] and y defaults to [j-1 for i=1:size(z,1),j=1:size(z,2)]

Examples

plot(cumsum(randn(100)))
plot(rand(5,5))

AsyPlots.process_pen_kwargs

AsyPlots.red

AsyPlots.sans_extension

AsyPlots.save — Function.

save(filename::AbstractString,
     P::Union{Plot2D,Plot3D};
     runasy=true,
     forcepdf=false)

Save Asymptote figure. If filename has extension .asy, then an asy file is saved together with any auxiliary data files.

If filename has extension .pdf, .svg or .png, then only the resulting image file is saved to the location filename

AsyPlots.sector

AsyPlots.showplots

AsyPlots.splitkwargs — Function.

splitkwargs(pooled_kwargs,options_to_separate)

Split a pooled list of keyword arguments into kwargs in options_to_separate and the rest

AsyPlots.updatedvals