Docstrings

EffectiveWaves.Acetone

EffectiveWaves.AirDry

EffectiveWaves.AverageWave — Type.

A type for the ensemble average scattering coefficients. Here they are discretised in terms of the depth x of the halfspace

EffectiveWaves.BS_matrices

EffectiveWaves.Benzene

EffectiveWaves.Brick

EffectiveWaves.Calcite

EffectiveWaves.Clay

EffectiveWaves.ClayRock

EffectiveWaves.EffectiveWave — Type.

A type for the effective wave ansatz.

EffectiveWaves.EffectiveWaves

EffectiveWaves.FusedSilica

EffectiveWaves.GlassPyrex

EffectiveWaves.Glycerol

EffectiveWaves.Hexadecane

EffectiveWaves.IronArmco

EffectiveWaves.LeadAnnealed

EffectiveWaves.LimeStone

EffectiveWaves.MatchWave — Type.

A type for matched waves.

EffectiveWaves.Medium — Type.

A homogenious background material.

EffectiveWaves.Nitrobenzene

EffectiveWaves.Nn

EffectiveWaves.OliveOil

EffectiveWaves.RubberGum

EffectiveWaves.SilicaQuartz

EffectiveWaves.SodiumNitrate

EffectiveWaves.Specie — Type.

A circular or spherical specie of particle with homogenious material properties.

EffectiveWaves.WaterDistilled

EffectiveWaves.Zn — Function.

Returns a ratio used in multiple scattering which reflects the scattering strength of the particles

EffectiveWaves.Zn_matrix — Function.

Pre-calculate a matrix of Zn's

EffectiveWaves.average_error — Function.

Approximates the error in AverageWave.amplitudes due to the mesh AverageWave.x.

EffectiveWaves.average_wave_system — Function.

note that this uses the non-dimensional X = k*depth

EffectiveWaves.d

EffectiveWaves.diff_far_field_pattern

EffectiveWaves.diffbesselj — Function.

Derivative of Bessel function of first kind

EffectiveWaves.diffhankelh1 — Function.

Derivative of Hankel function of the first kind

EffectiveWaves.effective_medium — Function.

the effective low frequency bulk modulus and density of a material filled with particles

EffectiveWaves.effective_waves — Function.

Calculates the effective wavenumbers and return Vector{EffectiveWave}.

EffectiveWaves.eval

EffectiveWaves.far_field_pattern

EffectiveWaves.gray_square

EffectiveWaves.gray_square!

EffectiveWaves.include

EffectiveWaves.integrate_B

EffectiveWaves.integrate_B_full

EffectiveWaves.integrate_S

EffectiveWaves.integration_scheme

EffectiveWaves.intergrand_kernel

EffectiveWaves.match_arrays — Function.

Returns (LT, ER, (imk^2invw).invVconj(wvec)), which connect the effective and average wave through α = LTA + (imk^2inv_w).invV*conj(w_vec). The matching region is X[L:end].

EffectiveWaves.match_error — Function.

Calculates the difference between the match of MatchWave.effectivewaves and MatchWave.averagewave. This can be used as a proxi for convergence.

EffectiveWaves.maximum_hankel_order — Function.

Calculate the largest needed order for the hankel series.

EffectiveWaves.one_species_low_wavenumber — Function.

low volfrac and low wavenumber. Fails badly for strong scatterers

EffectiveWaves.p_speed — Function.

Returns pressure wave speed, when β is the adiabatic bulk modulus/

EffectiveWaves.pair_field_pattern

EffectiveWaves.reduce_kvecs

EffectiveWaves.reduced_amplitudes_effective — Function.

The average effective transmitted scattering amplitudes for a given effective wavenumber keff. Assumes there exists only one keff. The function returns an array A, where AA(x,y,m,s) = im^mexp(-immθ_eff)A[m + maxhankelorder +1,s]exp(imkeff*(cos(θeff)x + sin(θin)y)) where (x,y) are coordinates in the halfspace, m-th hankel order, s-th species, and AA is the ensemble average scattering coefficient.

EffectiveWaves.reflection_coefficient — Function.

Calculates the reflection coefficient from each wave in waves and then sums the results.

The average reflection coefficient

EffectiveWaves.reflection_coefficient_halfspace — Function.

the reflection from a homogenious halfspace, which is also the low frequency reflection from a particulate material when using the effective_medium.

EffectiveWaves.reflection_coefficient_low_volfrac — Function.

An explicit formula for the refleciton coefficient based on a low particle volume fraction.

EffectiveWaves.reflection_coefficients — Function.

The average reflection coefficients

Pairs each ω in ωs with each wave in waves to calculate the refleciton coefficients: reflection_coefficient(ω, wave)

EffectiveWaves.scale_amplitudes_effective — Function.

returns a number a, such that a*As_eff will cancel an incident wave plane wave with incident angle θin.

EffectiveWaves.scattering_amplitudes_average

EffectiveWaves.simpson_scheme

EffectiveWaves.t_vectors — Function.

A t_matrix in the form of a vector, because for now we only deal with diagonal T matrices.

EffectiveWaves.transmission_angle — Function.

calculate effective transmission angle

EffectiveWaves.trap_scheme

EffectiveWaves.two_species_approx_wavenumber

EffectiveWaves.volume_fraction — Function.

Returns the volume fraction of the specie.

EffectiveWaves.wavenumber

EffectiveWaves.wavenumber_challis — Function.

wavenumber from Challis, R. E., et al. Ultrasound techniques for characterizing colloidal dispersions. Reports on progress in physics 68.7 (2005): 1541.

EffectiveWaves.wavenumber_far_field_low_volfrac

EffectiveWaves.wavenumber_low_volfrac — Function.

Explicit formula for one effective wavenumber based on a low particle volume fraction expasion.

EffectiveWaves.wavenumber_very_low_volfrac

EffectiveWaves.wavenumbers — Function.

Returns all the transmitted effective wavenumbers

EffectiveWaves.wavenumbers_mesh

EffectiveWaves.wavenumbers_path

EffectiveWaves.x_mesh — Function.

Returns x the mesh used to discretise the integral equations.

EffectiveWaves.x_mesh_match — Function.

Returns (x,L), where x[L:end] is the mesh used to match with wave_effs.

EffectiveWaves.ρ