Docstrings

NumericalIntegration.HALF

NumericalIntegration.IntegrationMethod

NumericalIntegration.NumericalIntegration

NumericalIntegration.RombergEven

NumericalIntegration.SimpsonEven

NumericalIntegration.SimpsonEvenFast

NumericalIntegration.Trapezoidal

NumericalIntegration.TrapezoidalEven

NumericalIntegration.TrapezoidalEvenFast

NumericalIntegration.TrapezoidalFast

NumericalIntegration._zero

NumericalIntegration._zeros

NumericalIntegration.cumul_integrate — Function.

cumul_integrate(x,y...)

Compute cumulative numerical integral of y(x) from x=x[1] to x=x[end]. Return a vector with elements of the same type as the input. If not method is supplied, use TrapezdoialFast.

cumul_integrate(x::AbstractVector, y::AbstractVector, ::Trapezoidal)

Use Trapezoidal rule.

cumul_integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalEven)

Use Trapezoidal rule, assuming evenly spaced vector x.

cumul_integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalFast)

Use Trapezoidal rule. Unsafe method: no bound checking. This is the default when no method is supplied.

cumul_integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalEvenFast)

Use Trapezoidal rule, assuming evenly spaced vector x. Unsafe method: no bound checking.

cumul_integrate(x::AbstractVector, y::AbstractMatrix, method; dims=2)

When y is an array, compute integral along each dimension specified by dims (default 2: columns)

NumericalIntegration.eval

NumericalIntegration.include

NumericalIntegration.integrate — Function.

integrate(x,y...)

Compute numerical integral of y(x) from x=x[1] to x=x[end]. Return a scalar of the same type as the input. If not method is supplied, use TrapezdoialFast.

integrate(x::AbstractVector, y::AbstractVector, ::Trapezoidal)

Use Trapezoidal rule.

integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalEven)

Use Trapezoidal rule, assuming evenly spaced vector x.

integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalFast)

Use Trapezoidal rule. Unsafe method: no bound checking. This is the default when no method is supplied.

integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalEvenFast)

Use Trapezoidal rule, assuming evenly spaced vector x. Unsafe method: no bound checking.

integrate(x::AbstractVector, y::AbstractVector, ::SimpsonEven)

Use Simpson's rule, assuming evenly spaced vector x.

integrate(x::AbstractVector, y::AbstractVector, ::SimpsonEven)

Use Simpson's rule, assuming evenly spaced vector x. Unsafe method: no bound checking.

integrate(x::AbstractVector, y::AbstractMatrix, method; dims=2)

When y is an array, compute integral along dimension specified by dims (default 2: columns).