Docstrings

Colors.@colorant_str — Macro.

Attempts to find a color to match any name you give it. First by using the methods from Color.jl's @colorant_str macro, to handle things like colorant"#cd32cd" Then if that fails, does a lookup in the big list of named colors. If that fails, then it provides suggestions for what color name you may have meant. As an user, you basically should just use this, without worrying if a color is defined. There are almost 3500 defined colors, and if you miss one, then the suggestions will help you out. (this is easier than trying to workout the name yourself)

NamedColors.ALL_COLORS

NamedColors.NamedColors

NamedColors.UnknownColorError

NamedColors.datafile

NamedColors.eval

NamedColors.include

NamedColors.load_all_colors — Function.

Loads a combined collection of all colors. Name clashes are resolved byt the load order. Which is defined, as from largest collection, to smalled collection. This is the initialised for the globel const ALL_COLORS, used by "@colorant"

NamedColors.load_crayola — Function.

Load the set of colors from Crayola Crayons. This is not the complete list, but it is extensive. These should be good for crayons

NamedColors.load_css3 — Function.

Load the W3C defined "Basic Colors". In theory these are as safe as it gets. These should be good everywhere.

NamedColors.load_nbs — Function.

Loads the NBS-ISCC color dictionary from

Kenneth L. Kelly and Deanne B. Judd.
"Color: Universal Language and Dictionary of Names",
National Bureau of Standards,
Spec. Publ. 440, Dec. 1976, 189 pages

Featuring the corrections by John Foster to the orignal computer entry of David Mundie

This is suitable for surface colors

NamedColors.load_resene — Function.

load the color list from Resene Paints Ltd These colors are best used with paint.

NamedColors.load_winsor_newton — Function.

Loads the Winsor-Newton Water-colours These should be good for surface colors

NamedColors.load_x11 — Function.

Loads the ``Standard'' X11 color-lists These colours are good for monitors. To be explict, these colours are good for Paul Raveling's HP Monitor

NamedColors.load_x11_style — Function.

Loads an x11 style color list. Lines are of the form r g b name which may have spaces the r, g and b terms are numbers between 0 and 256 in decimal Lines starting with ! are comments and are skipped

NamedColors.load_xcms_colordb — Function.

Load a XCMS color database See: https://www.x.org/archive/X11R7.5/doc/man/man3/XcmsColor.3.html

NamedColors.load_xkcd — Function.

Load the color list from the XKCD color survey. These colors are suitable for use with (modern) monitors

NamedColors.named_color — Function.

Returns a color with the given name. If not found,a list of suggestions is provided.

NamedColors.parse_CieLab — Function.

Parses strings in CIE lab, eg "2.03/-0.03/9.89"

NamedColors.parse_decimal256 — Function.

Parses three strings that are integers between 0-255, and converts to color

NamedColors.parse_hexcode — Function.

Parses as string of the form #RRGGBB eg "#32CD32" (limegreen) into a color

NamedColors.similarly_named_colors — Function.

Returns list of similarly named colors, from the color_list