Docstrings

Multisets.Multiset — Type.

A Multiset is an unordered collection of things with repetition permitted. A new Multiset container is created with Multiset{T}() where T is the type of the objects held in the multiset. If T is omitted, it defaults to Any.

A Multiset can be created from a collection list (such as a Vector or Set) with Multiset(list). If an element is repeated in list it has the appropriate multiplicity.

Multisets.Multisets

Multisets._inter

Multisets._minus

Multisets._sub

Multisets._union

Multisets.braces_string

Multisets.clean! — Function.

clean!(M) removes elements of multiplicy 0 from the underlying data structure supporting M.

Multisets.eval

Multisets.include

Multisets.julia_string

Multisets.multi_show_braces

Multisets.multi_show_flag

Multisets.multi_show_julia

Multisets.multi_show_short

Multisets.set_braces_show — Function.

Set braces display mode for multisets, like this:

{1,2,2,3,3,3,3}

See also set_short_show and set_julia_show.

Multisets.set_julia_show — Function.

Set Julia style display mode for multisets, like this:

Multiset(Int64[1,2,2,3,3,3,3])

See also set_short_show and set_braces_show.

Multisets.set_short_show — Function.

Set show display mode for multisets, like this:

Multiset{Int64} with 7 elements

See also set_braces_show and set_julia_show.

Multisets.short_string

Multisets.type_convert