Docstrings

BotCoreLCMTypes.BotCoreLCMTypes

BotCoreLCMTypes.atlas_command_t — Type.

modeled after the DRCSIM AtlasCommand ROS message, but updated for the atlas hardware API. control parameter spec:

q, qd, f are sensed position, velocity, torque, from AtlasJointState

qd, qdd, f_d are desired position, velocity, torque, from AtlasJointDesired

The final joint command will be: kqp * ( qd - q ) + kqi * 1/s * ( qd - q ) + kqdp * ( qdd - qd ) + kfp * ( fd - f ) + ffqd * qd + ffqdd * qdd + fffd * fd + ffconst

BotCoreLCMTypes.eval

BotCoreLCMTypes.force_torque_t

BotCoreLCMTypes.gps_data_t

BotCoreLCMTypes.gps_satellite_info_list_t

BotCoreLCMTypes.gps_satellite_info_t

BotCoreLCMTypes.image_metadata_t

BotCoreLCMTypes.image_sync_t

BotCoreLCMTypes.image_t

BotCoreLCMTypes.images_t

BotCoreLCMTypes.include

BotCoreLCMTypes.ins_t

BotCoreLCMTypes.joint_angles_t

BotCoreLCMTypes.joint_state_t

BotCoreLCMTypes.kvh_raw_imu_batch_t

BotCoreLCMTypes.kvh_raw_imu_t

BotCoreLCMTypes.planar_lidar_t

BotCoreLCMTypes.pointcloud2_t

BotCoreLCMTypes.pointcloud_t

BotCoreLCMTypes.pointfield_t

BotCoreLCMTypes.pose_t

BotCoreLCMTypes.position_3d_t

BotCoreLCMTypes.quaternion_t

BotCoreLCMTypes.raw_t

BotCoreLCMTypes.rigid_transform_t

BotCoreLCMTypes.robot_state_t

BotCoreLCMTypes.robot_urdf_t

BotCoreLCMTypes.sensor_status_t

BotCoreLCMTypes.six_axis_force_torque_array_t

BotCoreLCMTypes.six_axis_force_torque_t

BotCoreLCMTypes.system_status_t

BotCoreLCMTypes.twist_t

BotCoreLCMTypes.utime_t

BotCoreLCMTypes.vector_3d_t

BotCoreLCMTypes.viewer_command_t

BotCoreLCMTypes.viewer_draw_t

BotCoreLCMTypes.viewer_geometry_data_t

BotCoreLCMTypes.viewer_link_data_t

BotCoreLCMTypes.viewer_load_robot_t