Docstrings

PiCraft.AIR

PiCraft.BED

PiCraft.BEDROCK

PiCraft.BEDROCK_INVISIBLE

PiCraft.BOOKSHELF

PiCraft.BRICK_BLOCK

PiCraft.Block

PiCraft.CACTUS

PiCraft.CHEST

PiCraft.CLAY

PiCraft.COAL_ORE

PiCraft.COBBLESTONE

PiCraft.COBWEB

PiCraft.CRAFTING_TABLE

PiCraft.DIAMOND_BLOCK

PiCraft.DIAMOND_ORE

PiCraft.DIRT

PiCraft.DOOR_IRON

PiCraft.DOOR_WOOD

PiCraft.FARMLAND

PiCraft.FENCE

PiCraft.FENCE_GATE

PiCraft.FIRE

PiCraft.FLOWER_CYAN

PiCraft.FLOWER_YELLOW

PiCraft.FURNACE_ACTIVE

PiCraft.FURNACE_INACTIVE

PiCraft.GLASS

PiCraft.GLASS_PANE

PiCraft.GLOWING_OBSIDIAN

PiCraft.GLOWSTONE_BLOCK

PiCraft.GOLD_BLOCK

PiCraft.GOLD_ORE

PiCraft.GRASS

PiCraft.GRASS_TALL

PiCraft.GRAVEL

PiCraft.ICE

PiCraft.IRON_BLOCK

PiCraft.IRON_ORE

PiCraft.LADDER

PiCraft.LAPIS_LAZULI_BLOCK

PiCraft.LAPIS_LAZULI_ORE

PiCraft.LAVA

PiCraft.LAVA_FLOWING

PiCraft.LAVA_STATIONARY

PiCraft.LEAVES

PiCraft.MELON

PiCraft.MOSS_STONE

PiCraft.MUSHROOM_BROWN

PiCraft.MUSHROOM_RED

PiCraft.NBTag

PiCraft.NETHER_REACTOR_CORE

PiCraft.OBSIDIAN

PiCraft.PiCraft

PiCraft.REDSTONE_ORE

PiCraft.SAND

PiCraft.SANDSTONE

PiCraft.SAPLING

PiCraft.SNOW

PiCraft.SNOW_BLOCK

PiCraft.STAIRS_COBBLESTONE

PiCraft.STAIRS_WOOD

PiCraft.STONE

PiCraft.STONE_BRICK

PiCraft.STONE_SLAB

PiCraft.STONE_SLAB_DOUBLE

PiCraft.SUGAR_CANE

PiCraft.TAG_Byte

PiCraft.TAG_Byte_Array

PiCraft.TAG_Compound

PiCraft.TAG_Double

PiCraft.TAG_End

PiCraft.TAG_Float

PiCraft.TAG_Int

PiCraft.TAG_Int_Array

PiCraft.TAG_List

PiCraft.TAG_Long

PiCraft.TAG_Long_Array

PiCraft.TAG_Short

PiCraft.TAG_String

PiCraft.TNT

PiCraft.TORCH

PiCraft.WATER

PiCraft.WATER_FLOWING

PiCraft.WATER_STATIONARY

PiCraft.WOOD

PiCraft.WOOD_PLANKS

PiCraft.WOOL

PiCraft.World

PiCraft.__init__

PiCraft.buildModel — Function.

buildModel(m::Array{Block, 3}, pos::Tuple{Real, Real, Real})

Build a model from a 3-D block array.

PiCraft.camera

PiCraft.clearEvents — Function.

Clear all events in the buffer

PiCraft.connectToWorld — Function.

connectToWorld(address = "localhost", port = 4711)

Connect to the Minecraft API.

PiCraft.copyModel — Function.

copyModel(p1::Tuple{Real, Real, Real}, p2::Tuple{Real, Real, Real})

Copy the Blocks between the diagonally opposite blocks and store in a 3-D Block Array.

PiCraft.cutModel — Function.

cutModel(p1::Tuple{Real, Real, Real}, p2::Tuple{Real, Real, Real}, block::Block)

Copy blocks between diagonally opposite blocks and fill the space with block.

PiCraft.drawLine — Function.

drawLine(p1::Tuple{Real,Real,Real}, p2::Tuple{Real,Real,Real}, block::PiCraft.Block = PiCraft.GOLD_BLOCK, width::Real = 1)

Draw a line from point p1 to point p2 using block. block is gold block by default.

PiCraft.eval

PiCraft.exportSchematic

PiCraft.find_item

PiCraft.flip — Function.

flip(A::Array{Block, 3}, dim::Symbol)

Flip a 3-D Block Array in a particular dimension. Dimensions are x, y and z

Example

julia> A = reshape(Block.(collect(1:8)), (2,2,2))
2×2×2 Array{PiCraft.Block,3}:
[:, :, 1] =
 PiCraft.Block(1, 0)  PiCraft.Block(3, 0)
 PiCraft.Block(2, 0)  PiCraft.Block(4, 0)

[:, :, 2] =
 PiCraft.Block(5, 0)  PiCraft.Block(7, 0)
 PiCraft.Block(6, 0)  PiCraft.Block(8, 0)

julia> flip(A, :z)
2×2×2 Array{PiCraft.Block,3}:
[:, :, 1] =
 PiCraft.Block(5, 0)  PiCraft.Block(7, 0)
 PiCraft.Block(6, 0)  PiCraft.Block(8, 0)

[:, :, 2] =
 PiCraft.Block(1, 0)  PiCraft.Block(3, 0)
 PiCraft.Block(2, 0)  PiCraft.Block(4, 0)

PiCraft.getBlock — Function.

getBlock(pos::Tuple{Real, Real, Real})

Get the Block information from the specified coordinates.

PiCraft.getChatEvents

PiCraft.getHeight — Function.

getHeight(x::Int, z::Int)

Get the height of the world at the specified x and z coordinates.

PiCraft.getPlayerIds — Function.

getPlayerIds()

Return an array of all Player Id's connected to the server.

PiCraft.getPos — Function.

getPos()

Return the player's coordinates.

getPos(entityId::Int)

Get the position of the specified entity.

PiCraft.getTile — Function.

getTile()

Return the tile's coordinates on which the player is standing.

getTile(entityId::Int)

Get the tile on which the specified entity is.

PiCraft.importSchematic — Function.

importSchematic(filename::AbstractString)

Reads a .schematic file and return a 3 - D matrix of type Array{Block, 3}(model) which represents the schematic.

PiCraft.include

PiCraft.mc_items

PiCraft.mc_send — Function.

mc_send(cmd, output=true)

Communicate with the Minecraft API.

PiCraft.minecraftWorld

PiCraft.move — Function.

move(t::turtle, s::Int)

Move the turtle t forward s units.

PiCraft.parseNBT — Function.

parseNBT(filename::AbstractString, ostream::IO = stdout)

Parse a Named Binary tag file.

PiCraft.pitch — Function.

pitch(t::turtle, θ::Real)

Pitch the turtle t by θ degrees.

PiCraft.pollBlockHits — Function.

pollBlockHits()

Returns an array of all the events which have occurred since the last time the function was called. Each event is described with a tuple ((x, y, z), face, entityId). x, y and z are the coordinates of the block. face is the block's face number which was hit and entityId identifies the player who hit the block using a sword.

PiCraft.post — Function.

post(m...)

Post a message to chat

PiCraft.printTAG

PiCraft.readTAG — Function.

readTag(stream::IO, tagId = -1)

Read a single Binary Tag from stream. tagId is the type of nameless Tag tagId is -1 for Named Binary Tag.

PiCraft.restoreWorld — Function.

Restore the world to the previous savepoint.

PiCraft.roll — Function.

roll(t::turtle, θ::Real)

Roll the turtle t by θ degrees.

PiCraft.rotate — Function.

rotate(A::Array{Block, 3}; dir::Symbol = :l, dim::Symbol = :x)

Rotate a 3-D Block Array 90 degrees in a dimension and direction. Dimensions are x, y and z. Directions are l & r.

Example

julia> A = reshape(Block.(collect(1:8)), (2,2,2))
2×2×2 Array{PiCraft.Block,3}:
[:, :, 1] =
 PiCraft.Block(1, 0)  PiCraft.Block(3, 0)
 PiCraft.Block(2, 0)  PiCraft.Block(4, 0)

[:, :, 2] =
 PiCraft.Block(5, 0)  PiCraft.Block(7, 0)
 PiCraft.Block(6, 0)  PiCraft.Block(8, 0)

julia> rotate(A, dir = :r, dim = :y)
 2×2×2 Array{PiCraft.Block,3}:
 [:, :, 1] =
  PiCraft.Block(3, 0)  PiCraft.Block(7, 0)
  PiCraft.Block(4, 0)  PiCraft.Block(8, 0)

 [:, :, 2] =
  PiCraft.Block(1, 0)  PiCraft.Block(5, 0)
  PiCraft.Block(2, 0)  PiCraft.Block(6, 0)

PiCraft.rotate!

PiCraft.saveWorld — Function.

Save the World's progress

PiCraft.setBlock — Function.

setBlock(pos::Tuple{Real, Real, Real}, block::Block)

Place the specified Block at the given coordinates.

PiCraft.setBlocks — Function.

setBlocks(p1::Tuple{Real, Real, Real}, p2::Tuple{Real, Real, Real}, block::Block)

Set an entire region to the specified block type defined by the corners p1 and p2.

PiCraft.setPos — Function.

setPos(pos::Tuple{Real, Real, Real})

Teleport the player to the specified coordinates

setPos(entityId::Int, pos::Tuple{Real, Real, Real})

Set the position of the specified entity.

PiCraft.setTile — Function.

setTile(pos::Tuple{Real, Real, Real})

Teleport the player on top of the specified tile

setTile(entityId::Int)

Teleport the entity on the specified tile.

PiCraft.setting

PiCraft.tagDict

PiCraft.tagKeyDict

PiCraft.turtle — Type.

turtle

A graphics turtle meant to move in all 3-Dimensions. It contains the following fields:

pos : The coordinates of the turtle.

direction : The Roll Axis - direction in which the turtle is facing.

normal : The Yaw axis - direction normal to the turtle's body.

penBlock : The block which will be used to draw.

stepSize : The size of a turtle's step.

penDown : Activation state of penBlock.

The default constructor initializes the turtle at the player's position,roll axis is the positive x-direction, with the yaw axis pointing downwards. penBlock is gold by default and is activated with the stepSize = 0.5

PiCraft.writeTAG

PiCraft.yaw — Function.

yaw(t::turtle, θ::Real)

Yaw the turtle t by θ degrees.