Docstrings

Elly.APPLICATION_WEB_PROXY_BASE_ENV — Constant.

The environmental variable for APPLICATIONWEBPROXY_BASE. Set in ApplicationMaster's environment only. This states that for all non-relative web URLs in the app masters web UI what base should they have.

Elly.APP_CLASSPATH

Elly.APP_STATES — Constant.

APP_STATES: enum value to state map. Used for converting state for display.

Elly.APP_SUBMIT_TIME_ENV — Constant.

The environment variable for APPSUBMITTIME. Set in AppMaster environment only

Elly.ATTEMPT_STATES — Constant.

ATTEMPT_STATES: enum value to state map. Used for converting state for display.

Elly.CLASSPATH

Elly.CONTAINER_ID

Elly.CONTAINER_TOKEN_FILE_ENV_NAME — Constant.

The cache file into which container token is written

Elly.DEFAULT_FILE_MODE

Elly.DEFAULT_FOLDER_MODE

Elly.ELLY_CLIENTNAME

Elly.Elly

Elly.FINAL_APP_STATES — Constant.

FINALAPPSTATES: enum value to state map. Used for converting state for display.

Elly.HADOOP_COMMON_HOME

Elly.HADOOP_CONF_DIR

Elly.HADOOP_HDFS_HOME

Elly.HADOOP_YARN_HOME

Elly.HDATA_BLOCK_CHECKSUM

Elly.HDATA_COPY_BLOCK

Elly.HDATA_READ_BLOCK

Elly.HDATA_READ_METADATA

Elly.HDATA_REPLACE_BLOCK

Elly.HDATA_TRANSFER_BLOCK

Elly.HDATA_VERSION

Elly.HDATA_WRITE_BLOCK

Elly.HDFSBlockReader — Type.

HDFSBlockReader reads one block of data from a datanode.

Elly.HDFSBlockWriter — Type.

HDFSBlockWriter writes one block worth of data to a datanode.

The write method returns number of bytes accepted for writing into this block, which may be less than what was requested if block is full.

Packets are sent and acks received as and when data being written exceed packet size limit.

The close operation flushes remaining data in the block and waits for all pending acks.

Elly.HDFSClient — Type.

HDFSClient

A client to the namenode in a HDFS cluster and holds a connection to it. It also stores the folder context for using relative paths in APIs that use the client.

Elly.HDFSException — Type.

HDFSException

Thrown by HDFS APIs.

Elly.HDFSFile — Type.

HDFSFile

Holds a HDFSClient and path pointing to a file on HDFS

Elly.HDFSFileInfo — Type.

HDFSFileInfo

Stat structure for file/foledr in HDFS

Elly.HDFSFileReader — Type.

HDFSFileReader

Provides Julia IO APIs for reading HDFS files. Communicates with namenode for file metadata (through HDFSClient) and to datanodes for file data (through HDFSBlockReader)

Elly.HDFSFileWriter — Type.

HDFSFileWriter

Provides Julia IO APIs for writing HDFS files.

Steps:

get datanodes to write onto from namenode

break bytes into packets

write packets into first datanode's data channel (datanode will mirror packets to downstream datanodes)

when complete, confirm namenode with a blockReceived

On close call NameNode.complete to:

remove lease lease is a write lock for file modification. no leases are required for reading files. leases are managed by namenode lease is added on file create or append dfs client should start thread to renew leases periodically

change file from under construction to complete

Elly.HDFSProtocol — Type.

HDFSProtocol: Hadoop RPC client for namenode protocol.

Elly.HDFS_READER_WINDOW_LENGTH — Constant.

Default length (bytes) upto which to pre-fetch block metadata. (10 blocks of default size)

Elly.HOME

Elly.HRPC_AUTH_METHOD_SIMPLE

Elly.HRPC_AUTH_METHOD_TOKEN

Elly.HRPC_AUTH_PROTOCOL_NONE

Elly.HRPC_AUTH_PROTOCOL_SASL

Elly.HRPC_CALL_ID_CONNCTX

Elly.HRPC_CALL_ID_NORMAL

Elly.HRPC_CALL_ID_SASL

Elly.HRPC_FINAL_PACKET

Elly.HRPC_HEADER

Elly.HRPC_PROTOBUFF_TYPE

Elly.HRPC_PROTOCOLS

Elly.HRPC_SERVICE_CLASS

Elly.HRPC_VERSION

Elly.HadoopDataChannel — Type.

HadoopDataChannel is the connection to a datanode. Also associated with it is a buffer to help in message preparation.

Elly.HadoopDataChannelPool — Type.

HadoopDataChannelPool is a connection pool that holds connections to datanodes. Connections are deemed stale after keepalivesecs.

Elly.HadoopRpcChannel

Elly.HadoopRpcController

Elly.HadoopRpcException — Type.

HadoopRpcException is thrown on Rpc interaction errors either with namenode or datanode. Field status contains error code (enum) if received from the connected entity or just ERROR (1) to indicate failure. Though HadoopRpcException is used while communicating with both namenodes and datanodes, SUCCESS and ERROR are coded with the same values in both cases. Other specific error codes need knowledge of the context to interpret.

Elly.HadoopRpcProtocol — Type.

HadoopRpcProtocol binds a channel and controller with a service protocol implementation. Used by actual service implementations.

Elly.JAVA_HOME

Elly.LD_LIBRARY_PATH

Elly.LOCAL_DIRS

Elly.LOGNAME

Elly.LOG_DIRS

Elly.LOG_DIR_EXPANSION_VAR — Constant.

The temporary environmental variable for container log directory. This should be replaced by real container log directory on container launch.

Elly.Lock

Elly.MALLOC_ARENA_MAX

Elly.MAX_APP_ATTEMPTS_ENV — Constant.

The environment variable for MAXAPPATTEMPTS. Set in AppMaster environment only

Elly.NM_HOST

Elly.NM_HTTP_PORT

Elly.NM_PORT

Elly.NODE_STATES — Constant.

NODE_STATES: enum value to state map. Used for converting state for display.

Elly.PATH

Elly.PWD

Elly.PipelinedPacket — Type.

PipelinedPacket holds a packet of data while it is sent to datanodes and acknowledged. The data bytes are reset once acks are received successfully to reduce memory usage.

Elly.RequestPipeline — Type.

RequestPipeline holds entities while they are requested for from yarn resource manager. Application master thread extracts pending items and requests them from RM, whereupon they are moved on to the requested state.

Elly.SASL_DIGESTMD5_CHARSET

Elly.SASL_DIGESTMD5_MAXBUF

Elly.SASL_DIGESTMD5_METHOD

Elly.SASL_DIGESTMD5_NC

Elly.SHELL

Elly.TDCPoolElem

Elly.URI

Elly.USER

Elly.UserGroupInformation

Elly.WriterPipeline — Type.

WriterPipeline holds all packets of data for a block and provides methods to maintain their states. Pipeline failed status is set if any of the acks received is a failure. Field acked_bytes contains the count of bytes successfully sent till an error is encountered.

Elly.YARN_CONTAINER_CPU_DEFAULT

Elly.YARN_CONTAINER_LOCATION_DEFAULT

Elly.YARN_CONTAINER_MEM_DEFAULT

Elly.YARN_CONTAINER_PRIORITY_DEFAULT

Elly.YARN_NM_CONN_KEEPALIVE_SECS

Elly.YarnAMNMProtocol — Type.

YarnAMNMProtocol: Hadoop RPC client for Yarn application master to node manager protocol.

Elly.YarnAMRMProtocol — Type.

YarnAMRMProtocol: Hadoop RPC client for Yarn application master to resource manager protocol.

Elly.YarnApp — Type.

YarnApp represents one instance of application running on the yarn cluster

Elly.YarnAppAttemptStatus — Type.

YarnAppAttemptStatus wraps the protobuf type for ease of use

Elly.YarnAppMaster — Type.

YarnAppMaster is a skeleton application master. It provides the generic scafolding methods which can be used to create specific application masters for different purposes.

Elly.YarnAppStatus — Type.

YarnAppStatus wraps the protobuf type for ease of use

Elly.YarnClient — Type.

YarnClient holds a connection to the Yarn Resource Manager and provides APIs for application clients to interact with Yarn.

Elly.YarnClientProtocol — Type.

YarnClientProtocol: Hadoop RPC client for application client to Yarn resource manager protocol.

Elly.YarnContainers — Type.

YarnContainers holds all containers related to the application. It also holds the allocation and release pipelines that are used by application master for requesting actions from resource manager. Also schedules callbacks as tasks when containers are allocated or terminated.

Elly.YarnException — Type.

YarnException

Thrown by Yarn APIs.

Elly.YarnManager

Elly.YarnNode — Type.

YarnNode represents a node manager in the yarn cluster and its communication address, resource state and run state.

Elly.YarnNodes — Type.

YarnNodes holds node information as visible to the application master. It also caches connection to node masters. Connection are reused if they are required before a set keepalivesecs time.

Elly._a

Elly._add_block

Elly._as_dict

Elly._block_has_offset

Elly._can_write

Elly._complete_file

Elly._container

Elly._create_file

Elly._current_window

Elly._currprocname

Elly._dcpool

Elly._envdict

Elly._find_block

Elly._get

Elly._get_block_locations

Elly._get_content_summary

Elly._get_file_info

Elly._get_fs_status

Elly._get_server_defaults

Elly._len_uleb

Elly._new_app

Elly._new_or_existing_conn

Elly._put

Elly._read_and_buffer

Elly._select_node_for_block

Elly._unregister

Elly._update_rm

Elly._walkdir

Elly._write

Elly.abspath

Elly.ackrcvd

Elly.ackwait

Elly.add_token

Elly.am_rm_token

Elly.assert_no_failure

Elly.atime

Elly.attempts

Elly.begin_send

Elly.buffer_client_read_status

Elly.buffer_conctx_reqhdr

Elly.buffer_connctx

Elly.buffer_handshake

Elly.buffer_method_reqhdr

Elly.buffer_opcode

Elly.buffer_readblock

Elly.buffer_reqhdr

Elly.buffer_rpc_reqhdr

Elly.buffer_sasl_message

Elly.buffer_sasl_reqhdr

Elly.buffer_size_delimited

Elly.buffer_writeblock

Elly.bytesavailable

Elly.calc_chksum

Elly.callback

Elly.can_schedule

Elly.can_schedule_cores

Elly.can_schedule_mem

Elly.cd

Elly.check_write_buffer

Elly.close

Elly.conditional_sasl_auth

Elly.container_allocate

Elly.container_release

Elly.container_start

Elly.container_stop

Elly.convert

Elly.cp

Elly.digmd5_challenge_isvalid

Elly.digmd5_decode_challenge

Elly.digmd5_respond

Elly.dirname

Elly.disconnect

Elly.du

Elly.enqueue

Elly.eof

Elly.eofpacket

Elly.eval

Elly.exists

Elly.filemode

Elly.filesize

Elly.find_tokens

Elly.flush

Elly.flush_and_wait

Elly.get_connection

Elly.hadoop

Elly.haverequests

Elly.hdfs_blocks

Elly.hdfs_capacity

Elly.hdfs_capacity_remaining

Elly.hdfs_capacity_used

Elly.hdfs_default_block_size

Elly.hdfs_default_replication

Elly.hdfs_renewlease — Function.

Applications that write infrequently/slowly must call renewlease periodically to prevent the namenode from assuming the client from having abandoned the file or some other client from recovering the lease.

Elly.hdfs_server_defaults

Elly.hdfs_set_replication

Elly.hdfs_status

Elly.hdr_iob

Elly.include

Elly.initiate

Elly.isconnected

Elly.isdir

Elly.isfile

Elly.islink

Elly.isvalid_chksum

Elly.joinpath

Elly.kill

Elly.launch

Elly.launchcontext

Elly.makelock

Elly.manage

Elly.md5

Elly.mkdir

Elly.mkpath

Elly.move

Elly.mtime

Elly.mv

Elly.next_call_id

Elly.nodecount

Elly.nodes

Elly.open

Elly.pending

Elly.populate_checksums

Elly.position

Elly.prepare_packet

Elly.process_am_rm

Elly.process_pipeline

Elly.pwd

Elly.read

Elly.read!

Elly.read_packet! — Function.

Read one packet into inbuff starting from offset. If inbuff has insufficient space, returns the minimum additional space required in inbuff to read the packet as a negative number. Otherwise, returns the number of bytes available in inbuff after reading the packet.

Elly.read_packet_ack

Elly.readall

Elly.readbytes

Elly.readdir

Elly.recv_blockop

Elly.recv_packet_hdr

Elly.recv_rpc_message

Elly.recv_sasl_message

Elly.register

Elly.release_connection

Elly.renewlease

Elly.request_alloc

Elly.request_release

Elly.rm

Elly.sasl_auth

Elly.seek

Elly.seekend

Elly.seekstart

Elly.send_block_read

Elly.send_block_write

Elly.send_buffered

Elly.send_read_status

Elly.send_rpc_message

Elly.set_busy

Elly.set_debug

Elly.set_free

Elly.setup_worker

Elly.show

Elly.skip

Elly.stat

Elly.status

Elly.submit

Elly.testdigmd5

Elly.token_alias

Elly.torequest

Elly.touch

Elly.trigger_flushed

Elly.trigger_pkt

Elly.unregister

Elly.update

Elly.username

Elly.verify_pkt_checksums

Elly.wait_flushed

Elly.wait_for_attempt_state

Elly.wait_for_state

Elly.wait_pkt

Elly.withlock

Elly.write

Elly.write_packet