Docstrings

StrAPI.ASCIICompare

StrAPI.All

StrAPI.AlwaysValid

StrAPI.ByteCompare

StrAPI.CharSetASCIICompat — Type.

Characters 0-0x7f same as ASCII

StrAPI.CharSetBMP — Type.

Characters 0-0xd7ff, 0xe000-0xffff follow Unicode BMP

StrAPI.CharSetBinary — Type.

8-bit Binary string, not text

StrAPI.CharSetISOCompat — Type.

Characters 0-0x9f follows ISO 8859

StrAPI.CharSetOther — Type.

Codepoints are not in Unicode compatible order

StrAPI.CharSetStyle — Type.

CharSetStyle(A)
CharSetStyle(typeof(A))

CharSetStyle specifies the information about the character set used by the string or characters. When you define a new AbstractString or AbstractChar type, you can choose to implement it with

StrAPI.CharSetStyle(::Type{<:MyString}) = StrAPI.CharSetISOCompat()

The default is CharSetUnicode()

StrAPI.CharSetUnicode — Type.

Full Unicode character set, no additions

StrAPI.CharSetUnicodePlus — Type.

Unicode character set, plus encodings of invalid characters

StrAPI.CharSetUnknown — Type.

Raw bytes, words, or character string, unknown encoding/character set

StrAPI.CodePointCompare

StrAPI.CodePoints

StrAPI.CodeUnitTypes

StrAPI.CompareStyle — Type.

CompareStyle(Union{A, typeof(A)}, Union{B, typeof(B)})

CompareStyle specifies how to compare two strings with character set encodings A and B

StrAPI.CompareStyle(::Type{<:MyString}, ::Type{String}) = StrAPI.ByteCompare()

The default is CodePointCompare

StrAPI.Direction

StrAPI.Each

StrAPI.EncodingStyle — Type.

EncodingStyle(A)
EncodingStyle(typeof(A))

EncodingStyle specifies the whether a character set encoding uses one or multiple codeunits to encode a single codepoint. When you define a new AbstractString type, you can choose to implement it with either single or multi-codeunit indexing.

StrAPI.EncodingStyle(::Type{MyCharSetEncoding}) = StrAPI.MultiCodeUnitEncoding()

The default is SingleCodeUnitEncoding()

StrAPI.FindOp

StrAPI.First

StrAPI.Fwd

StrAPI.Last

StrAPI.MaybeSub

StrAPI.MultiCodeUnitEncoding

StrAPI.NEW_ITERATE

StrAPI.Next

StrAPI.NoCompare

StrAPI.Prev

StrAPI.Rev

StrAPI.SingleCodeUnitEncoding

StrAPI.StrAPI — Module.

StrAPI package

Copyright 2017-2018 Gandalf Software, Inc., Scott P. Jones Licensed under MIT License, see LICENSE.md

StrAPI.StrErrors — Module.

Based partly on code in LegacyStrings that used to be part of Julia Licensed under MIT License, see LICENSE.md

(Written by Scott P. Jones in series of PRs contributed to the Julia project in 2015)

StrAPI.StringError

StrAPI.UC

StrAPI.UTF16Compare

StrAPI.Uni — Module.

Unicode Normalization, Category constants and documentation

Copyright 2017-2018 Gandalf Software, Inc. Scott P. Jones, and other contributors to the UTF8proc library (for the documentation of the API)

Licensed under MIT License, see LICENSE.md

StrAPI.Unicode

StrAPI.UnknownValidity

StrAPI.ValidatedStyle — Type.

ValidatedStyle(A)
ValidatedStyle(typeof(A))

ValidatedStyle specifies the whether a string or character type is always valid or not When you define a new AbstractString or AbstractChar type, you can choose to implement it as always validated, or validation state unknown.

StrAPI.ValidatedStyle(::Type{<:MyCharType}) = StrAPI.AlwaysValid()

The default is UnknownValidity

StrAPI.WidenCompare

StrAPI.WordCompare

StrAPI.__api__

StrAPI.__tmp_api__

StrAPI._isvalid

StrAPI._lowercase

StrAPI._print

StrAPI._repmsg

StrAPI._titlecase

StrAPI._uppercase

StrAPI._write

StrAPI.argerror

StrAPI.ascii_err

StrAPI.basetype — Function.

Get the base type (of CodeUnitTypes) of a character or aligned/swapped type

StrAPI.boundserr

StrAPI.category_abbrev

StrAPI.category_code

StrAPI.category_string

StrAPI.charset — Function.

Get the character set used by a string or character type

StrAPI.chr2ind

StrAPI.codepoint_error

StrAPI.codepoints — Function.

codepoints(str)

An iterator that generates the code points of a string

Examples

julia> a = str("abc🖖")

julia> collect(a)

julia> collect(codepoints(a))

StrAPI.create_vector

StrAPI.cse — Function.

Get the character set / encoding used by a string type

StrAPI.curmod — Module.

StrAPI package

Copyright 2017-2018 Gandalf Software, Inc., Scott P. Jones Licensed under MIT License, see LICENSE.md

StrAPI.encoding — Function.

Get the character set used by a string type

StrAPI.ends_with

StrAPI.eval

StrAPI.find

StrAPI.find_result

StrAPI.fnd

StrAPI.found

StrAPI.get_iobuffer

StrAPI.graphemes

StrAPI.include

StrAPI.ind2chr

StrAPI.index_error

StrAPI.is_alphabetic

StrAPI.is_alphanumeric

StrAPI.is_ascii

StrAPI.is_assigned

StrAPI.is_bmp

StrAPI.is_control

StrAPI.is_defined

StrAPI.is_digit

StrAPI.is_empty

StrAPI.is_grapheme_break

StrAPI.is_grapheme_break!

StrAPI.is_graphic

StrAPI.is_hex_digit

StrAPI.is_latin

StrAPI.is_letter

StrAPI.is_lowercase

StrAPI.is_multi — Function.

Determine if a string has multiple codeunit encoding

StrAPI.is_numeric

StrAPI.is_printable

StrAPI.is_punctuation

StrAPI.is_space

StrAPI.is_unicode

StrAPI.is_uppercase

StrAPI.is_valid

StrAPI.lowercase_first

StrAPI.namlst

StrAPI.ncharerr

StrAPI.neginderr

StrAPI.nulerr

StrAPI.occurs_in

StrAPI.outhex

StrAPI.pr_ul

StrAPI.pwc

StrAPI.quotesym

StrAPI.repeaterr

StrAPI.starts_with

StrAPI.str_done

StrAPI.str_next

StrAPI.strerror

StrAPI.symstr

StrAPI.text_width

StrAPI.uppercase_first

StrAPI.utf8crc