Docstrings

Pukeko.@parametric — Macro.

@parametric func iterable

Create a version of func that is prefixed with TEST_PREFIX in the module that this macro is called for each value in iterable. If a value in iterable is a tuple, it is splatted into the function arguments.

Pukeko.@test — Macro.

@test expression

Test that expression is true.

Pukeko.@test_throws — Macro.

@test_throws exception_type expression

Test that expression throws an exception of type exception_type.

Pukeko.Pukeko

Pukeko.TEST_PREFIX — Constant.

TEST_PREFIX

Functions with this string at the the start of their name will be treated as self-contained sets of tests.

Pukeko.TestException — Type.

TestException

The Exceptionthrown when a Pukeko test fails. Used byrun_tests` to distinguish between test errors and unexpected errors.

Pukeko.eval

Pukeko.include

Pukeko.parametric — Function.

parametric(module_to_test, func, iterable)

Create a version of func that is prefixed with TEST_PREFIX in module_to_test for each value in iterable. If a value in iterable is a tuple, it is splatted into the function arguments.

Pukeko.run_tests — Function.

run_tests(module_to_test; fail_fast=false)

Runs all the sets of tests in module module_to_test. Test sets are defined as functions with names that begin with TEST_PREFIX. A summary is printed after all test sets have been run and if there were any failures an exception is thrown.

Configuration options:

If fail_fast==false (default), if any one test function fails, the others will still run. If true, testing will stop on the first failure. The commandline argument --PUKEKO_FAIL_FAST will override fail_fast to true for all run_tests calls.

If timing==true (default is false), print elapsed time and memory allocation statistics for every test function. The commandline argument --PUKEKO_TIMING will override timing to true for all run_tests calls.

If match_name!=nothing (default is nothing), only run tests that contain match_name in their names. The commandline argument --PUKEKO_MATH=str will override match_name to str for all run_tests calls.

Pukeko.test_equal — Function.

test_equal(value_left, value_right, [linfo])

Test that value_left is equal to value_right. Calls to this are generated by @test for the case of @test expr_left == expr_right.

Pukeko.test_true — Function.

test_true(value, [linfo])

Throws iff value is not true. Calls to this are generated by @test.