Docstrings

FunctionalTables.ColumnSort — Type.

struct ColumnSort{K, R}

Sort specification for a column. K::Symbol is a key for sorting, R::Bool reverses sorting for this key.

This type is not part of the interface, for internal representation.

FunctionalTables.ColumnSorting — Type.

Type for sorting, used internally.

FunctionalTables.FunctionalTable — Type.

FunctionalTable(itr)
FunctionalTable(itr, sortspecs)
FunctionalTable(itr, sortspecs, sortingpolicy; cfg)

Create a FunctionalTable from an iterable that returns NamedTuples.

Returned values need to have the same names (but not necessarily types).

sorting specifies sorting, and is a tuple of :key or :key => reverse elements.

cfg determines sink configuration for collecting elements of the columns, see SinkConfig.

FunctionalTables.FunctionalTables

FunctionalTables.Keys — Type.

Type for keys, used internally.

FunctionalTables.RLEVector — Type.

struct RLEVector{C, T, S}

An RLE encoded vector, using negative lengths for missing values.

When an elemenet in counts is positive, it encodes that many of the corresponding element in data.

Negative counts encode missing values of type S (has to be a concrete singleton). In this case there is no corresponding value in data, ie data may have fewer elements than counts.

An RLEVector can also act as a column.

FunctionalTables.RepeatedValue — Type.

RepeatedValue(value, len)

Equivalent to a vector containing len instances of value. Used internally.

FunctionalTables.SHOWROWS — Constant.

Shows this many values from each column in a FunctionalTable.

FunctionalTables.SINKCONFIG — Constant.

Default sink configuration.

FunctionalTables.SINKVECTORS — Constant.

Sink configuration that collects to vectors.

FunctionalTables.SinkConfig — Type.

struct SinkConfig{M}

FunctionalTables.SortingPolicy — Type.

struct SortingPolicy{K}

Policy for dealing with specified sortings.

See VerifySorting, TrustSorting, and TrySorting.

FunctionalTables.SplitTable — Type.

struct SplitTable{K, T<:FunctionalTable, C<:SinkConfig}

Implements by.

Iterator state is a tuple, with

sinks and firstkey, created from the element with a non-matching key,

itrstate, the iteration state for itr.

FunctionalTables.TrustSorting — Type.

Accept the specified sorting to hold without any checks (except for verifying that column names are valid).

Note

This can lead to incorrect results, use cautiously. VerifySorting is recommended instead as it has little overhead.

FunctionalTables.TrySorting — Type.

Try the specified sorting, then gracefully degrade to a subset of it that holds.

FunctionalTables.VerifySorting — Type.

Verify that the specified sorting holds. This is the default sorting policy.

FunctionalTables._cmp_sorting

FunctionalTables._collect_block!

FunctionalTables._retained_sorting

FunctionalTables._select_sorting

FunctionalTables._showcolcontents

FunctionalTables.append1 — Function.

append1(v, elt)

Append elt to v, allocating a new vector and copying the contents.

Type of new collection is calculated using promote_type.

FunctionalTables.by — Function.

by(indexkeys, ft; cfg)

An iterator that groups rows of tables by the columns indexkeys, returning (index::NamedTupe, table::FunctionalTable) for each contiguous block of the index keys.

cfg is used for collecting table.

FunctionalTables.cancontain — Function.

cancontain(T, elt)

Test if a collection of element type T can contain a new element elt without any loss of precision.

FunctionalTables.checkvalidkeys — Function.

checkvalidkeys(argkeys, validkeys)

Check that argkeys are a subset of the valid keys.

When that does not hold, throw and error with an informative message.

FunctionalTables.cmp_columnsort

FunctionalTables.cmp_sorting — Function.

cmp_sorting(cs, a, b)

Compare rows a and b, which support the getproperty interface, with the given column sorting.

Internal.

FunctionalTables.collect_column — Function.

collect_column(cfg, itr)

Collect results from itr into a sink (using config cfg), then finalize and return the column.

FunctionalTables.collect_column!

FunctionalTables.collect_columns — Function.

columns, sorting =

collect_columns(cfg, itr, sorting, sortingpolicy)

Collect results from itr, which are supposed to be NamedTuples with the same names, into sinks (using config cfg), then finalize and return the NamedTuple of the columns and the sorting (which is ≡ sorting unless sortingpolicy ≡ SortingPolicy(:prefix)).

sorting is ignored when sortingpolicy ≡ SortingPolicy(:accept), otherwise used according to the value of the latter.

FunctionalTables.collect_columns!

FunctionalTables.column_sorting — Function.

column_sorting(sortspecs)
column_sorting(sortspecs, colkeys)

Process sorting specifications for columns (an iterable or possibly a ColumnSorting), return a value of type ColumnSorting.

Verify that sort keys are unique. When colkeys is given, verify that the sort keys are a subset of it.

Accepted syntax:

:key, for sorting a column in ascending order

:key => reverse, for sorting a column in descending order

All functions which accept sort specs should use this, but the function itself is not part of the API.

FunctionalTables.columns — Function.

columns(ft; vector, mutable)

Return the columns in a NamedTuple.

When mutable, all columns will be mutable <: AbstractVector, and not share (shallow) structure.

When vector, all columns will be <: AbstractVector, but may be immutable or share structure.

FunctionalTables.contiguous_blocks

FunctionalTables.dropkeys — Function.

dropkeys(ftkeys, drop)

Check that drop ⊆ ftkeys, then return ftkeys ∖ drop.

FunctionalTables.eval

FunctionalTables.finalize_sink — Function.

finalize_sink(?, sink)

Convert sink to a column.

sink may share structure with the result and is not supposed to be used for saving any more elements.

FunctionalTables.finalize_sinks — Function.

finalize_sinks(cfg, sinks)

Finalize a (named) tuple of sinks.

FunctionalTables.fuse — Function.

fuse(f, index, tables)

Prepend the index as repeated columns to f(index, tables...).

FunctionalTables.getsorting

FunctionalTables.include

FunctionalTables.isless_sorting

FunctionalTables.make_sink — Function.

make_sink(cfg, elt)

Create and return a sink using configuration cfg that stores elt.

FunctionalTables.make_sinks — Function.

make_sinks(cfg, elts)

Make sinks for a (named) tuple pf elements.

FunctionalTables.merge_repeated — Function.

merge_repeated(index, ft)

Make a functional table from index, repeating each value for a column to match the length of ft, then merge the two.

FunctionalTables.merge_sorting — Function.

merge_sorting(cs, otherkeys)

Calculate sorting when a table with sorting is merged with a table containing otherkeys, which may replace columns.

FunctionalTables.narrow — Function.

narrow(x)

Convert the argument to a narrower type if possible without losing precision.

Note

This function is not type stable, use only when new container types are determined.

FunctionalTables.retained_sorting — Function.

retained_sorting(cs, a, b)

Return the part of sorting under which a ≤ b.

FunctionalTables.select — Function.

select(ft, keep)

select(ft, keep...) select(ft; drop)

Select a subset of columns from the table.

select(ft, keep) and select(ft, keep...) returns the table with the given columns.

select(ft; drop = keys) is a convenience form for keeping all but the given columns.

FunctionalTables.select_sorting — Function.

select_sorting(cs, keep)

Calculate sorting when only keep keys are kept.

keep may contain keys not in the sorting, ie those of a FunctionalTable.

FunctionalTables.sort_in_memory

FunctionalTables.sortkey — Function.

Accessor for sort key. Internal.

FunctionalTables.split_namedtuple — Function.

split_namedtuple(splitter, nt)

Splits a named tuple in two, based on the names in splitter.

Returns two NamedTuples; the first one is ordered as splitter, the second one with the remaining values as in the original argument.

julia> split_namedtuple(NamedTuple{(:a, :c)}, (c = 1, b = 2, a = 3, d = 4))
((a = 3, c = 1), (b = 2, d = 4))

FunctionalTables.store!_or_reallocate — Function.

store!_or_reallocate(?, sink, elt)

Either store elt in sink (in which case the returned value is ≡ sink), or allocate a new sink that can do this, copy the contents, save elt and return that (then the returned value is ≢ sink).