Docstrings

DocSeeker.CACHE

DocSeeker.CACHETIMEOUT

DocSeeker.DOCDBCACHE

DocSeeker.DocObj

DocSeeker.DocSeeker

DocSeeker.MAX_RETURN_SIZE

DocSeeker.__init

DocSeeker._createdocsdb

DocSeeker._loaddocsdb

DocSeeker.alldocs — Function.

alldocs() -> Vector{DocObj}

Find all docstrings in all currently loaded Modules.

DocSeeker.baseURL — Function.

baseURL(links::Vector{Markdown.Link}) -> String

Find the most common host and return the first URL in links with that host.

DocSeeker.basepath

DocSeeker.createdocsdb — Function.

createdocsdb()

Asynchronously create a "database" of all local docstrings in Pkg.dir(). This is done by loading all packages and using introspection to retrieve the docstrings – the obvious limitation is that only packages that actually load without errors are considered.

DocSeeker.dbpath

DocSeeker.determinetype

DocSeeker.docsdir — Function.

docsdir(pkg) -> String

Find the directory conataining the documenatation for package pkg. Will fall back to returning a package's README.md. Returns an empty String if no indication of documentation is found.

DocSeeker.docsurl — Function.

docsurl(pkg) -> String

Return the most likely candidate for a package's online documentation or an empty string.

DocSeeker.dynamicsearch

DocSeeker.eval

DocSeeker.finddocsURL — Function.

finddocsURL(pkg) -> Vector{Markdown.Link}

Search pkgs readme for links to documentation.

DocSeeker.findlinks

DocSeeker.findplainlinks

DocSeeker.getfiles

DocSeeker.include

DocSeeker.isdoclink

DocSeeker.jlhome

DocSeeker.keywords

DocSeeker.load_documenter_docs

DocSeeker.loaddocsdb — Function.

loaddocsdb() -> Vector{DocObj}

Retrieve the docstrings from the "database" created by createdocsdb(). Will return an empty vector if the database is locked by createdocsdb().

DocSeeker.lockpath

DocSeeker.modulebindings

DocSeeker.pkgrootpath

DocSeeker.readmepath

DocSeeker.score — Function.

score(needle, s::Docs.DocObj) -> Float

Scores s against the search query needle. Returns a Float between 0 and 1.

DocSeeker.searchdocs

DocSeeker.searchfiles