Docstrings

SimpleRandom.Bernoulli_RV — Function.

Bernoulli(p) makes a single coin flip RV.

SimpleRandom.Binomial_RV — Function.

Binomial_RV(n,p) returns a binomial random variable.

SimpleRandom.E — Function.

E(X) is the expected value of X.

SimpleRandom.RV — Type.

RV represents a discrete random variable with finite support.

SimpleRandom.RV_types

SimpleRandom.SimpleRandom

SimpleRandom.Uniform_RV — Function.

Uniform_RV(n) returns the uniform distribution on {1,2,...,n}.

SimpleRandom.Var — Function.

Var(Y) is the variance of Y.

SimpleRandom.binom_rv — Function.

binom_rv(n,p) generates a random binomial random value. p defaults to 0.5.

SimpleRandom.eval

SimpleRandom.exp_rv — Function.

exp_rv(theta) returns an exponential random value with mean theta (which defaults to 1.0).

SimpleRandom.include

SimpleRandom.poisson_rv — Function.

poisson_rv(lambda) generates a Poisson random value with mean lamba (which defaults to 1.0).

SimpleRandom.probs — Function.

probs(X::RV) returns an iterator of the probabilities associated with the values in X.

SimpleRandom.random_choice — Function.

random_choice(X) for a random variable X returns a value of X according to its probability distribution. That is, the probability a value v is returned is X[v].

random_choice(weights) randomly chooses a value from 1 to n, where n is the number of elements in weights. The probability that k is chosen is proportional to weights[k]. The weights must be nonnegative and not all zero.

random_choice(dict) choose a random key k from dict with weight proportional to dict[k]. Thus, dict must be of type Dict{S, T<:Real}.

SimpleRandom.random_subset — Function.

random_subset is used to create random subsets as follows:

random_subset(A): random subset of A with each element

chosen with probability 1/2.

random_subset(A,k): random k-element subset of A.

random_subset(n): random subset of 1:n.

random_subset(n,k): random k-element subset of 1:n.

SimpleRandom.random_unit_vector — Function.

random_unit_vector(d) returns a random d-dimensional unit vector.

SimpleRandom.report — Function.

report(X) prints out a list of the values of X and their associated probabilities

SimpleRandom.validate! — Function.

validate!(X) ensures that the probabilies of the values in X sum to one. If not, they are rescaled.

SimpleRandom.vals — Function.

vals(X::RV) returns an iterator of the values this random variable can take. Use X[v] to get the associate probability of the value v.