Docstrings

Rematch.@match — Macro.

Patterns:

_ matches anything

foo matches anything, binds value to foo

Foo(x,y,z) matches structs of type Foo with fields matching x,y,z

Foo(y=1) matches structs of type Foo whose y field equals 1

[x,y,z] matches AbstractArrays with 3 entries matching x,y,z

(x,y,z) matches Tuples with 3 entries matching x,y,z

[x,y...,z] matches AbstractArrays with at least 2 entries, where x matches the first entry, z matches the last entry and y matches the remaining entries.

(x,y...,z) matches Tuples with at least 2 entries, where x matches the first entry, z matches the last entry and y matches the remaining entries.

_::T matches any subtype (isa) of T

x || y matches values which match either x or y (only variables which exist in both branches will be bound)

x && y matches values which match both x and y

x where condition matches only if condition is true (condition may use any variables that occur earlier in the pattern eg (x, y, z where x + y > z))

Anything else is treated as a constant and tested for equality

Expressions can be interpolated in as constants via standard interpolation syntax $(x)

Patterns can be nested arbitrarily.

Repeated variables only match if they are equal (==). For example (x,x) matches (1,1) but not (1,2).

Rematch.@splice

Rematch.MatchFailure

Rematch.Rematch

Rematch.assert_num_fields

Rematch.eval

Rematch.evaluated_fieldcount

Rematch.handle_destruct

Rematch.handle_destruct_fields

Rematch.handle_match_case

Rematch.handle_match_cases

Rematch.handle_match_eq

Rematch.include