Docstrings

ModuleInterfaceTools.@api — Macro.

@api <cmd> [<symbols>...]

freeze # use at end of module to freeze API

list <modules>... # list API(s) of given modules (or current if none given)

use <modules>... # use, without importing (i.e. can't extend)

use! <modules>... # use, without importing (i.e. can't extend), "export"

test <modules>... # using public and develop APIs, for testing purposes

extend <modules>... # for development, imports api & dev, use api & dev definitions

extend! <modules>... # for development, imports api & dev, use api & dev definitions, "export"

export <modules>... # export public definitions

base <names...> # Add functions from Base that are part of the API (extendible)

base! <names...> # Add functions from Base or define them if not in Base

public <names...> # Add other symbols that are part of the public API (structs, consts)

public! <names...> # Add functions that are part of the public API (extendible)

develop <names...> # Add other symbols that are part of the development API

develop! <names...> # Add functions that are part of the development API (extendible)

define! <names...> # Define functions to be extended, public API

defdev! <names...> # Define functions to be extended, develop API

modules <names...> # Add submodule names that are part of the API

path <paths...> # Add paths to LOAD_PATH

def <name> <expr> # Same as the @def macro, creates a macro with the given name

ModuleInterfaceTools.API

ModuleInterfaceTools.AbstractAPI

ModuleInterfaceTools.BIG_ENDIAN

ModuleInterfaceTools.ModuleInterfaceTools — Module.

API Tools package

Copyright 2018 Gandalf Software, Inc., Scott P. Jones

Licensed under MIT License, see LICENSE.md

(@def macro "stolen" from DiffEqBase.jl/src/util.jl :-) )

ModuleInterfaceTools.SymList

ModuleInterfaceTools.SymSet

ModuleInterfaceTools.TMP_API

ModuleInterfaceTools.V6_COMPAT

ModuleInterfaceTools._add_def!

ModuleInterfaceTools._add_symbols — Function.

Add symbols

ModuleInterfaceTools._api

ModuleInterfaceTools._api_def

ModuleInterfaceTools._api_display

ModuleInterfaceTools._api_export

ModuleInterfaceTools._api_extend

ModuleInterfaceTools._api_freeze

ModuleInterfaceTools._api_list

ModuleInterfaceTools._api_path

ModuleInterfaceTools._api_test

ModuleInterfaceTools._api_use

ModuleInterfaceTools._cmdadd

ModuleInterfaceTools._cmduse

ModuleInterfaceTools._do_list

ModuleInterfaceTools._ff

ModuleInterfaceTools._init_api — Function.

Initialize the temp api variable for this module

ModuleInterfaceTools._stderr

ModuleInterfaceTools._stdout

ModuleInterfaceTools.debug

ModuleInterfaceTools.eval

ModuleInterfaceTools.get_api

ModuleInterfaceTools.has_api

ModuleInterfaceTools.include

ModuleInterfaceTools.m_eval

ModuleInterfaceTools.makecmd

ModuleInterfaceTools.push_args!

ModuleInterfaceTools.showeval