Docstrings

PCRE2.ALLOW_EMPTY_CLASS

PCRE2.ALT_BSUX

PCRE2.ALT_CIRCUMFLEX

PCRE2.ALT_VERBNAMES

PCRE2.ANCHORED

PCRE2.AUTO_CALLOUT

PCRE2.BSR_ANYCRLF

PCRE2.BSR_UNICODE

PCRE2.CASELESS

PCRE2.COMMON_MASK

PCRE2.COMPILE_MASK

PCRE2.CONFIG

PCRE2.CONFIG_BSR

PCRE2.CONFIG_COMPILED_WIDTHS

PCRE2.CONFIG_DEPTHLIMIT

PCRE2.CONFIG_HEAPLIMIT

PCRE2.CONFIG_JIT

PCRE2.CONFIG_JITTARGET

PCRE2.CONFIG_LINKSIZE

PCRE2.CONFIG_MATCHLIMIT

PCRE2.CONFIG_NEVER_BACKSLASH_C

PCRE2.CONFIG_NEWLINE

PCRE2.CONFIG_PARENSLIMIT

PCRE2.CONFIG_STACKRECURSE

PCRE2.CONFIG_UNICODE

PCRE2.CONFIG_UNICODE_VERSION

PCRE2.CONFIG_VERSION

PCRE2.CONVERT_GLOB

PCRE2.CONVERT_GLOB_NO_STARSTAR

PCRE2.CONVERT_GLOB_NO_WILD_SEPARATOR

PCRE2.CONVERT_NO_UTF_CHECK

PCRE2.CONVERT_POSIX_BASIC

PCRE2.CONVERT_POSIX_EXTENDED

PCRE2.CONVERT_UTF

PCRE2.CodeP

PCRE2.CodeUnitTypes

PCRE2.CompileContextP

PCRE2.CompileOptions

PCRE2.ConvertContextP

PCRE2.DFA_RESTART

PCRE2.DFA_SHORTEST

PCRE2.DOLLAR_ENDONLY

PCRE2.DOTALL

PCRE2.DUPNAMES

PCRE2.ENDANCHORED

PCRE2.EXTENDED

PCRE2.EXTENDED_MORE

PCRE2.EXTRA_ALLOW_SURROGATE_ESCAPES

PCRE2.EXTRA_BAD_ESCAPE_IS_LITERAL

PCRE2.EXTRA_MATCH_LINE

PCRE2.EXTRA_MATCH_WORD

PCRE2.FIRSTLINE

PCRE2.GeneralContextP

PCRE2.INFO

PCRE2.INFO_ALLOPTIONS

PCRE2.INFO_ARGOPTIONS

PCRE2.INFO_BACKREFMAX

PCRE2.INFO_BSR

PCRE2.INFO_CAPTURECOUNT

PCRE2.INFO_DEPTHLIMIT

PCRE2.INFO_EXTRAOPTIONS

PCRE2.INFO_FIRSTBITMAP

PCRE2.INFO_FIRSTCODETYPE

PCRE2.INFO_FIRSTCODEUNIT

PCRE2.INFO_FRAMESIZE

PCRE2.INFO_HASBACKSLASHC

PCRE2.INFO_HASCRORLF

PCRE2.INFO_HEAPLIMIT

PCRE2.INFO_JCHANGED

PCRE2.INFO_JITSIZE

PCRE2.INFO_LASTCODETYPE

PCRE2.INFO_LASTCODEUNIT

PCRE2.INFO_MATCHEMPTY

PCRE2.INFO_MATCHLIMIT

PCRE2.INFO_MAXLOOKBEHIND

PCRE2.INFO_MINLENGTH

PCRE2.INFO_NAMECOUNT

PCRE2.INFO_NAMEENTRYSIZE

PCRE2.INFO_NAMETABLE

PCRE2.INFO_NEWLINE

PCRE2.INFO_SIZE

PCRE2.JIT_COMPLETE

PCRE2.JIT_PARTIAL_HARD

PCRE2.JIT_PARTIAL_SOFT

PCRE2.JitStackP

PCRE2.LITERAL

PCRE2.MATCH_MASK

PCRE2.MATCH_UNSET_BACKREF

PCRE2.MULTILINE

PCRE2.MatchContextP

PCRE2.MatchDataP

PCRE2.MatchOptions

PCRE2.NEVER_BACKSLASH_C

PCRE2.NEVER_UCP

PCRE2.NEVER_UTF

PCRE2.NEWLINE_ANY

PCRE2.NEWLINE_ANYCRLF

PCRE2.NEWLINE_CR

PCRE2.NEWLINE_CRLF

PCRE2.NEWLINE_LF

PCRE2.NEWLINE_NUL

PCRE2.NL_MASK

PCRE2.NOTBOL

PCRE2.NOTEMPTY

PCRE2.NOTEMPTY_ATSTART

PCRE2.NOTEOL

PCRE2.NO_AUTO_CAPTURE

PCRE2.NO_AUTO_POSSESS

PCRE2.NO_DOTSTAR_ANCHOR

PCRE2.NO_JIT

PCRE2.NO_START_OPTIMIZE

PCRE2.NO_UTF_CHECK

PCRE2.PARTIAL_HARD

PCRE2.PARTIAL_SOFT

PCRE2.PCRE2

PCRE2.PCRE2_Error

PCRE2.SUBSTITUTE_EXTENDED

PCRE2.SUBSTITUTE_GLOBAL

PCRE2.SUBSTITUTE_OVERFLOW_LENGTH

PCRE2.SUBSTITUTE_UNKNOWN_UNSET

PCRE2.SUBSTITUTE_UNSET_EMPTY

PCRE2.SizeP

PCRE2.SizeRef

PCRE2.StrP

PCRE2.UCP

PCRE2.UCharP

PCRE2.UNGREEDY

PCRE2.UNSET

PCRE2.USE_OFFSET_LIMIT

PCRE2.UTF

PCRE2.V6_COMPAT

PCRE2.VoidP

PCRE2._UCHAR

PCRE2.__init__

PCRE2._ncodeunits

PCRE2.capture_names

PCRE2.check_deps

PCRE2.code_copy

PCRE2.code_copy_with_tables

PCRE2.code_free

PCRE2.compile — Function.

Wrapper for compile() function, throw error on error return with error info

PCRE2.compile_context_copy

PCRE2.compile_context_create

PCRE2.compile_context_free

PCRE2.compile_error

PCRE2.config

PCRE2.convert_context_copy

PCRE2.convert_context_create

PCRE2.convert_context_free

PCRE2.converted_pattern_free

PCRE2.create_vector

PCRE2.depsjl_path

PCRE2.dfa_match

PCRE2.err_message

PCRE2.eval

PCRE2.evr

PCRE2.funclist

PCRE2.general_context_copy

PCRE2.general_context_free

PCRE2.get_error_message

PCRE2.get_mark

PCRE2.get_ovec

PCRE2.get_ovector_count

PCRE2.get_ovector_pointer

PCRE2.get_startchar

PCRE2.include

PCRE2.info — Function.

PCRE2 pattern_info call wrapper

PCRE2.info_error

PCRE2.jit_compile — Function.

Wrapper for jit_compile() function, throw error on error return with error info

PCRE2.jit_error

PCRE2.jit_free_unused_memory

PCRE2.jit_match

PCRE2.jit_stack_assign

PCRE2.jit_stack_create

PCRE2.jit_stack_free

PCRE2.libpcre2_16

PCRE2.libpcre2_32

PCRE2.libpcre2_8

PCRE2.maketables

PCRE2.match

PCRE2.match_context_copy

PCRE2.match_context_create

PCRE2.match_context_free

PCRE2.match_data_create

PCRE2.match_data_create_from_pattern

PCRE2.match_data_free

PCRE2.pattern_convert

PCRE2.pattern_info

PCRE2.pcre_error

PCRE2.serialize_decode

PCRE2.serialize_encode

PCRE2.serialize_free

PCRE2.serialize_get_number_of_codes

PCRE2.set_bsr

PCRE2.set_character_tables

PCRE2.set_compile_extra_options

PCRE2.set_depth_limit

PCRE2.set_glob_escape

PCRE2.set_glob_separator

PCRE2.set_heap_limit

PCRE2.set_match_limit

PCRE2.set_max_pattern_length

PCRE2.set_newline

PCRE2.set_offset_limit

PCRE2.set_parens_nest_limit

PCRE2.set_recursion_limit

PCRE2.sub_copy_bynumber

PCRE2.sub_length_bynumber

PCRE2.substitute

PCRE2.substring_copy_byname

PCRE2.substring_copy_bynumber

PCRE2.substring_free

PCRE2.substring_get_byname

PCRE2.substring_get_bynumber

PCRE2.substring_length_byname

PCRE2.substring_length_bynumber

PCRE2.substring_list_free

PCRE2.substring_list_get

PCRE2.substring_nametable_scan

PCRE2.substring_number_from_name