Docstrings

Tables.ColumnTable

Tables.ColumnsRow

Tables.EachColumn

Tables.IteratorInterfaceExtensions

Tables.IteratorRow

Tables.IteratorWrapper

Tables.MatrixRow

Tables.MatrixTable

Tables.NamedTupleIterator

Tables.Row — Type.

Abstract row type with a simple required interface: row values are accessible via getproperty(row, field); for example, a NamedTuple like nt = (a=1, b=2, c=3) can access its value for a like nt.a which turns into a call to the function getproperty(nt, :a)

Tables.RowIterator

Tables.RowTable

Tables.Schema — Type.

Tables.Schema(names, types)

Create a Tables.Schema object that holds the column names and types for a tabular data object. Tables.Schema is dual-purposed: provide an easy interface for users to query these properties, as well as provide a convenient "structural" type for code generation.

To get a table's schema, one can call Tables.schema(tbl), but also note that a table may return nothing, indicating that it's column names and/or column types are unknown (usually not inferrable). This is similar to the Base.EltypeUnknown() trait for iterators when Base.IteratorEltype is called. Users should account for the Tables.schema(tbl) => nothing case by using the properties of the results of Tables.rows(x) and Tables.columns(x) directly.

To access the names, one can simply call sch.names to return the tuple of Symbols. To access column types, one can similarly call sch.types, which will return a tuple of types (like (Int64, Float64, String)).

The actual type definition is

struct Schema{names, types} end

Where names is a tuple of Symbols, and types is a tuple type of types (like Tuple{Int64, Float64, String}). Encoding the names & types as type parameters allows convenient use of the type in generated functions and other optimization use-cases.

Tables.Select

Tables.SelectRow

Tables.Table — Type.

The Tables.jl package provides simple, yet powerful interface functions for working with all kinds tabular data through predictable access patterns.

    Tables.rows(table) => Rows
    Tables.columns(table) => Columns

Where Rows and Columns are the duals of each other:

Rows is an iterator of property-accessible objects (any type that supports propertynames(row) and getproperty(row, nm::Symbol)

Columns is a property-accessible object of iterators (i.e. each column is an iterator)

In addition to these Rows and Columns objects, it's useful to be able to query properties of these objects:

Tables.schema(x::Union{Rows, Columns}) => Union{Tables.Schema, Nothing}: returns a Tables.Schema object, or nothing if the table's schema is unknown

For the Tables.Schema object:

column names can be accessed as a tuple of Symbols like sch.names

column types can be accessed as a tuple of types like sch.types

See ?Tables.Schema for more details on this type

A big part of the power in these simple interface functions is that each (Tables.rows & Tables.columns) is defined for any table type, even if the table type only explicitly implements one interface function or the other. This is accomplished by providing performant, generic fallback definitions in Tables.jl itself (though obviously nothing prevents a table type from implementing each interface function directly).

With these simple definitions, powerful workflows are enabled:

A package providing data cleansing, manipulation, visualization, or analysis can automatically handle any number of decoupled input table types

A tabular file format can have automatic integration with in-memory structures and translation to other file formats

So how does one go about satisfying the Tables.jl interface functions? It mainly depends on what you've already defined and the natural access patterns of your table:

First:

Tables.istable(::Type{<:MyTable}) = true: this provides an explicit affirmation that your type implements the Tables interface

To support Rows:

Define Tables.rowaccess(::Type{<:MyTable}) = true: this signals to other types that MyTable supports valid Row-iteration

Define Tables.rows(x::MyTable): return a Row-iterator object (perhaps the table itself if already defined)

Define Tables.schema(Tables.rows(x::MyTable)) to either return a Tables.Schema object, or nothing if the schema is unknown or non-inferrable for some reason

To support Columns:

Define Tables.columnaccess(::Type{<:MyTable}) = true: this signals to other types that MyTable supports returning a valid Columns object

Define Tables.columns(x::MyTable): return a Columns, property-accessible object (perhaps the table itself if it naturally supports property-access to columns)

Define Tables.schema(Tables.columns(x::MyTable)) to either return a Tables.Schema object, or nothing if the schema is unknown or non-inferrable for some reason

The final question is how MyTable can be a "sink" for any other table type. The answer is quite simple: use the interface functions!

Define a function or constructor that takes, at a minimum, a single, untyped argument and then calls Tables.rows or Tables.columns on that argument to construct an instance of MyTable

For example, if MyTable is a row-oriented format, I might define my "sink" function like:

function MyTable(x)
    Tables.istable(x) || throw(ArgumentError("MyTable requires a table input"))
    rows = Tables.rows(x)
    sch = Tables.schema(rows)
    names = sch.names
    types = sch.types
    # custom constructor that creates an "empty" MyTable according to given column names & types
    # note that the "unknown" schema case should be considered, i.e. when `sch.types => nothing`
    mytbl = MyTable(names, types)
    for row in rows
        # a convenience function provided in Tables.jl for "unrolling" access to each column/property of a `Row`
        # it works by applying a provided function to each value; see `?Tables.eachcolumn` for more details
        Tables.eachcolumn(sch, row) do val, col, name
            push!(mytbl[col], val)
        end
    end
    return mytbl
end

Alternatively, if MyTable is column-oriented, perhaps my definition would be more like:

function MyTable(x)
    Tables.istable(x) || throw(ArgumentError("MyTable requires a table input"))
    cols = Tables.columns(x)
    # here we use Tables.eachcolumn to iterate over each column in a `Columns` object
    return MyTable(collect(propertynames(cols)), [collect(col) for col in Tables.eachcolumn(cols)])
end

Obviously every table type is different, but via a combination of Tables.rows and Tables.columns each table type should be able to construct an instance of itself.

Tables.Tables

Tables.Transforms

Tables.TransformsRow

Tables.__init__

Tables._buildcolumns

Tables._eltype

Tables._types

Tables.add!

Tables.add_or_widen!

Tables.allocatecolumn — Function.

Tables.allocatecolumn(::Type{T}, len) => returns a column type (usually AbstractVector) w/ size to hold `len` elements

Custom column types can override with an appropriate "scalar" element type that should dispatch to their column allocator.

Tables.allocatecolumns

Tables.buildcolumns

Tables.columnaccess

Tables.columnindex — Function.

given names and a Symbol name, compute the index (1-based) of the name in names

Tables.columns

Tables.columntable

Tables.columntype — Function.

given tuple type and a Symbol name, compute the type of the name in the tuples types

Tables.ctappend

Tables.datavaluetype

Tables.eachcolumn — Function.

Tables.eachcolumn(f, sch, row, args...)
Tables.eachcolumn(f, names, row, args...)
Tables.eachcolumn(Tables.columns(x))

The first two definitions take a function `f`, table schema `sch` or table's column names `names` as a tuple of Symbols, a `row` type (that satisfies the Row interface), and any other `args...`;
they generate calls to get the value for each column in the row (`getproperty(row, nm)`) and then calls `f(val, col, name, args...)`, where `f` is the
user-provided function, `val` is a row's column value, `col` is the column index as an `Int`, and `name` is the row's column name as a Symbol.

While the first two definitions apply to a `Row` object, the last definition simply returns a property-iterator over a `Columns` object.
For example, one could "collect" every column of a `Columns` object by doing:
```julia
vectors = [collect(col) for col in Tables.eachcolumn(Tables.columns(x))]
```

Tables.eval

Tables.getarray

Tables.getfunc

Tables.haslength

Tables.include

Tables.invalidtable

Tables.istable

Tables.materializer

Tables.matrix — Function.

Tables.matrix(table)

Materialize any table source input as a Matrix. If the table column types are not homogenous, they will be promoted to a common type in the materialized Matrix. Note that column names are ignored in the conversion.

Tables.namedtupleiterator

Tables.names

Tables.nondatavaluetype

Tables.rowaccess

Tables.rowcount

Tables.rows

Tables.rowtable

Tables.runlength — Function.

helper function to calculate a run-length encoding of a tuple type

Tables.schema

Tables.select

Tables.table — Function.

Tables.table(m::AbstractMatrix; [header::Vector{Symbol}])

Wrap an AbstractMatrix (Matrix, Adjoint, etc.) in a MatrixTable, which satisfies the Tables.jl interface. This allows accesing the matrix via Tables.rows and Tables.columns. An optional keyword argument header can be passed as a Vector{Symbol} to be used as the column names. Note that no copy of the AbstractMatrix is made.

Tables.transform

Tables.types

Tables.typesubset

Tables.unweakref

Tables.unweakreftype

Tables.unweakreftypes

Tables.unwrap