Docstrings

DoubleFloats.@d16_str

DoubleFloats.@d32_str

DoubleFloats.@d64_str

DoubleFloats.AccurateDWMinusDW

DoubleFloats.AccurateDWPlusDW

DoubleFloats.BigFloatBits

DoubleFloats.BigFloatDigits

DoubleFloats.DWDivDW2

DoubleFloats.DWDivDW3

DoubleFloats.DWDivFP3

DoubleFloats.DWInvDW2

DoubleFloats.DWInvDW3

DoubleFloats.DWMinusFP

DoubleFloats.DWPlusFP

DoubleFloats.DWTimesDW2

DoubleFloats.DWTimesDW3

DoubleFloats.DWTimesFP1

DoubleFloats.DWTimesFP3

DoubleFloats.DomainGuard — Type.

DomainGuard{T}

An easy way to provide mathematical functions guards that disallow out-of-domain processing describes the canonical value-span and offers the manner of reductive domain comprehension.

For discontinous [gapping} periodic functions, all that spans any of a subdomain collective is included and admissable. Each evaluand is provisional; quantities with qualities map the domain-reachable onto canonical values.

DoubleFloats.Double16

DoubleFloats.Double32

DoubleFloats.Double64

DoubleFloats.DoubleFloat

DoubleFloats.DoubleFloats

DoubleFloats.Fast2Mult

DoubleFloats.Fast2Sum

DoubleFloats.FloatSigned

DoubleFloats.HI

DoubleFloats.HILO

DoubleFloats.LO

DoubleFloats.MultipartFloat

DoubleFloats.QuadrupleF16

DoubleFloats.QuadrupleF32

DoubleFloats.QuadrupleF64

DoubleFloats.TangentCoeffs

DoubleFloats.TwoDiff

DoubleFloats.TwoSum

DoubleFloats.abs_db_db

DoubleFloats.abs_dd_dd

DoubleFloats.abs_fp_dd

DoubleFloats.abs_fpdb

DoubleFloats.add133

DoubleFloats.add2 — Function.

add2(x::T, y::T) where {T<:AbstractFloat}

Add x and y as if they are DoubleFloat{T} numbers. This is more efficient as first converting to DoubleFloat{T} and then adding.

DoubleFloats.add222

DoubleFloats.add223

DoubleFloats.add313

DoubleFloats.add322

DoubleFloats.add323

DoubleFloats.add332

DoubleFloats.add333

DoubleFloats.add_dbdb_db

DoubleFloats.add_dbdb_db_nonfinite

DoubleFloats.add_dbfp_db

DoubleFloats.add_dbfp_db_nonfinite

DoubleFloats.add_dddd_dd

DoubleFloats.add_ddfp_dd

DoubleFloats.add_fpdb_db

DoubleFloats.add_fpdb_db_nonfinite

DoubleFloats.add_fpdd_dd

DoubleFloats.add_fpfp_db

DoubleFloats.add_fpfp_dd

DoubleFloats.bernoulli_even

DoubleFloats.calc_exp

DoubleFloats.cbrt_db_db

DoubleFloats.cbrt_dd_dd

DoubleFloats.cos_circle

DoubleFloats.cos_kernel

DoubleFloats.cos_npio32

DoubleFloats.cos_pi1o1024_t64

DoubleFloats.cos_pi1o128_t64

DoubleFloats.cos_pi1o12_t64

DoubleFloats.cos_pi1o15_t64

DoubleFloats.cos_pi1o16_t64

DoubleFloats.cos_pi1o2048_t64

DoubleFloats.cos_pi1o24_t64

DoubleFloats.cos_pi1o36_t64

DoubleFloats.cos_pi1o3_t64

DoubleFloats.cos_pi1o48_t64

DoubleFloats.cos_pi1o4_t64

DoubleFloats.cos_pi1o512_t64

DoubleFloats.cos_pi1o5_t64

DoubleFloats.cos_pi1o64_t64

DoubleFloats.cos_pi1o6_t64

DoubleFloats.cos_pi1o7_t64

DoubleFloats.cos_pi1o8_t64

DoubleFloats.cos_taylor

DoubleFloats.cot_pi1o1024_t64

DoubleFloats.cot_pi1o128_t64

DoubleFloats.cot_pi1o12_t64

DoubleFloats.cot_pi1o15_t64

DoubleFloats.cot_pi1o16_t64

DoubleFloats.cot_pi1o2048_t64

DoubleFloats.cot_pi1o24_t64

DoubleFloats.cot_pi1o36_t64

DoubleFloats.cot_pi1o3_t64

DoubleFloats.cot_pi1o48_t64

DoubleFloats.cot_pi1o4_t64

DoubleFloats.cot_pi1o512_t64

DoubleFloats.cot_pi1o5_t64

DoubleFloats.cot_pi1o64_t64

DoubleFloats.cot_pi1o6_t64

DoubleFloats.cot_pi1o7_t64

DoubleFloats.cot_pi1o8_t64

DoubleFloats.cot_taylor

DoubleFloats.csc_pi1o1024_t64

DoubleFloats.csc_pi1o128_t64

DoubleFloats.csc_pi1o12_t64

DoubleFloats.csc_pi1o15_t64

DoubleFloats.csc_pi1o16_t64

DoubleFloats.csc_pi1o2048_t64

DoubleFloats.csc_pi1o24_t64

DoubleFloats.csc_pi1o36_t64

DoubleFloats.csc_pi1o3_t64

DoubleFloats.csc_pi1o48_t64

DoubleFloats.csc_pi1o4_t64

DoubleFloats.csc_pi1o512_t64

DoubleFloats.csc_pi1o5_t64

DoubleFloats.csc_pi1o64_t64

DoubleFloats.csc_pi1o6_t64

DoubleFloats.csc_pi1o7_t64

DoubleFloats.csc_pi1o8_t64

DoubleFloats.csc_taylor

DoubleFloats.cube

DoubleFloats.cube_db_db

DoubleFloats.cube_dd_dd

DoubleFloats.cube_fp_dd

DoubleFloats.div2 — Function.

div(x::T, y::T) where {T<:AbstractFloat}

Divide x and y as if they are DoubleFloat{T} numbers. This is more efficient as first converting to DoubleFloat{T} and then dividing.

DoubleFloats.double

DoubleFloats.double16

DoubleFloats.double16inv

DoubleFloats.double32

DoubleFloats.double32inv

DoubleFloats.double64

DoubleFloats.double64inv

DoubleFloats.double_inv

DoubleFloats.dvi_dbdb_db

DoubleFloats.dvi_dbdb_db_nonfinite

DoubleFloats.dvi_dbfp_db

DoubleFloats.dvi_dbfp_db_nonfinite

DoubleFloats.dvi_dddd_dd

DoubleFloats.dvi_dddd_dd_fast

DoubleFloats.dvi_ddfp_dd

DoubleFloats.dvi_fpdb_db

DoubleFloats.dvi_fpdb_db_nonfinite

DoubleFloats.dvi_fpdd_dd

DoubleFloats.dvi_fpdd_dd_fast

DoubleFloats.dvi_fpfp_db

DoubleFloats.dvi_fpfp_dd

DoubleFloats.eval

DoubleFloats.exp_half_one

DoubleFloats.exp_int

DoubleFloats.exp_taylor

DoubleFloats.exp_zero_half

DoubleFloats.exponent

DoubleFloats.fma_

DoubleFloats.fma_2

DoubleFloats.fma_3

DoubleFloats.fma_4

DoubleFloats.fmod

DoubleFloats.fracpart

DoubleFloats.halfpi

DoubleFloats.halfpi_d64

DoubleFloats.halfpi_df64

DoubleFloats.halfpi_minus

DoubleFloats.hash_0_double_lo

DoubleFloats.hash_double_lo

DoubleFloats.inc_pi_1o1024_t64

DoubleFloats.include

DoubleFloats.index_npio32

DoubleFloats.inf

DoubleFloats.intpart

DoubleFloats.inv_db_db

DoubleFloats.inv_dd_dd

DoubleFloats.inv_dd_dd_fast

DoubleFloats.inv_fact

DoubleFloats.inv_fp_db

DoubleFloats.inv_fp_dd

DoubleFloats.inv_halfpi_d64

DoubleFloats.inv_onepi_d64

DoubleFloats.inv_pi_15o32_d64

DoubleFloats.inv_pi_15o32_t64

DoubleFloats.inv_pi_1o128_d64

DoubleFloats.inv_pi_1o128_t64

DoubleFloats.inv_pi_1o12_d64

DoubleFloats.inv_pi_1o12_t64

DoubleFloats.inv_pi_1o16_d64

DoubleFloats.inv_pi_1o16_t64

DoubleFloats.inv_pi_1o1_d64

DoubleFloats.inv_pi_1o1_t64

DoubleFloats.inv_pi_1o2048_t64

DoubleFloats.inv_pi_1o24_d64

DoubleFloats.inv_pi_1o24_t64

DoubleFloats.inv_pi_1o256_d64

DoubleFloats.inv_pi_1o256_t64

DoubleFloats.inv_pi_1o2_d64

DoubleFloats.inv_pi_1o2_t64

DoubleFloats.inv_pi_1o32_d64

DoubleFloats.inv_pi_1o32_t64

DoubleFloats.inv_pi_1o3_d64

DoubleFloats.inv_pi_1o3_t64

DoubleFloats.inv_pi_1o4_d64

DoubleFloats.inv_pi_1o4_t64

DoubleFloats.inv_pi_1o512_t64

DoubleFloats.inv_pi_1o64_d64

DoubleFloats.inv_pi_1o64_t64

DoubleFloats.inv_pi_1o6_d64

DoubleFloats.inv_pi_1o6_t64

DoubleFloats.inv_pi_1o8_d64

DoubleFloats.inv_pi_1o8_t64

DoubleFloats.inv_pi_2o1_d64

DoubleFloats.inv_pi_2o1_t64

DoubleFloats.inv_pi_4o1_d64

DoubleFloats.inv_pi_4o1_t64

DoubleFloats.inv_pi_7o16_d64

DoubleFloats.inv_pi_7o16_t64

DoubleFloats.inv_qrtrpi_d64

DoubleFloats.inv_twopi_d64

DoubleFloats.isfractional

DoubleFloats.isneg

DoubleFloats.isneginf

DoubleFloats.isnonneg

DoubleFloats.isnonpos

DoubleFloats.isnonzero

DoubleFloats.ispos

DoubleFloats.isposinf

DoubleFloats.k_pio2

DoubleFloats.logten

DoubleFloats.logtwo

DoubleFloats.minus_halfpi

DoubleFloats.minus_pi

DoubleFloats.minus_qrtrpi

DoubleFloats.minus_twopi

DoubleFloats.mod2pi_neg

DoubleFloats.modhalfpi

DoubleFloats.modhalfpi_neg

DoubleFloats.modhalfpipm

DoubleFloats.modpi

DoubleFloats.modpi_neg

DoubleFloats.modpipm

DoubleFloats.modqrtrpi

DoubleFloats.modqrtrpi_neg

DoubleFloats.modqrtrpipm

DoubleFloats.mul132

DoubleFloats.mul133

DoubleFloats.mul2 — Function.

mul2(x::T, y::T) where {T<:AbstractFloat}

Multiply x and y as if they are DoubleFloat{T} numbers. This is more efficient as first converting to DoubleFloat{T} and then multiplying.

DoubleFloats.mul222 — Function.

mul222( (ahi, alo), (bhi, blo) ) ↦ (chi, clo)

Preconditions

abs(alo) ≤ abs(ahi) * pow(2, -53)

abs(blo) ≤ abs(bhi) * pow(2, -53)

Postconditions

abs(clo) ≤ abs(ahi) * pow(2, -53)

chi + clo == ((ahi + alo) * (bhi + blo)) * (1 + ε), abs(ε) ≤ pow(2, -102)

chi + clo == chi

TripleDouble_BuildingBlocks

DoubleFloats.mul223

DoubleFloats.mul232

DoubleFloats.mul233

DoubleFloats.mul312

DoubleFloats.mul313

DoubleFloats.mul322

DoubleFloats.mul323

DoubleFloats.mul332

DoubleFloats.mul333

DoubleFloats.mul_by_half

DoubleFloats.mul_by_two

DoubleFloats.mul_dbdb_db

DoubleFloats.mul_dbdb_db_nonfinite

DoubleFloats.mul_dbfp_db

DoubleFloats.mul_dbfp_db_nonfinite

DoubleFloats.mul_dddd_dd

DoubleFloats.mul_ddfp_dd

DoubleFloats.mul_fpdb_db

DoubleFloats.mul_fpdb_db_nonfinite

DoubleFloats.mul_fpdd_dd

DoubleFloats.mul_fpfp_db

DoubleFloats.mul_fpfp_dd

DoubleFloats.mul_pow2

DoubleFloats.mul_pwr2

DoubleFloats.n_bernoulli_even

DoubleFloats.n_cos_npio32

DoubleFloats.n_exp_int

DoubleFloats.n_npio32

DoubleFloats.n_sin_npio32

DoubleFloats.n_tan_coeff

DoubleFloats.n_tan_npio32

DoubleFloats.nan

DoubleFloats.neg_db_db

DoubleFloats.neg_dd_dd

DoubleFloats.neg_fp_db

DoubleFloats.neg_fp_dd

DoubleFloats.negabs

DoubleFloats.negabs_db_db

DoubleFloats.negabs_dd_dd

DoubleFloats.negabs_fp_db

DoubleFloats.negabs_fp_dd

DoubleFloats.neginf

DoubleFloats.ninv_fact

DoubleFloats.normalize — Function.

normalize(x,y)

x_normalized^2 + y_normalized^2 == one(promote_type(typeof(x),typeof(y)))

DoubleFloats.npio32

DoubleFloats.nused_inv_fact

DoubleFloats.one_d64

DoubleFloats.one_t64

DoubleFloats.onepi

DoubleFloats.onepi_d64

DoubleFloats.onepi_df64

DoubleFloats.pi_15o32_d64

DoubleFloats.pi_15o32_t64

DoubleFloats.pi_1o1024_t64

DoubleFloats.pi_1o128_d64

DoubleFloats.pi_1o128_t64

DoubleFloats.pi_1o12_d64

DoubleFloats.pi_1o12_t64

DoubleFloats.pi_1o16_d64

DoubleFloats.pi_1o16_t64

DoubleFloats.pi_1o1_d64

DoubleFloats.pi_1o1_t64

DoubleFloats.pi_1o1_t64_hi

DoubleFloats.pi_1o1_t64_lo

DoubleFloats.pi_1o1_t64_md

DoubleFloats.pi_1o2048_t64

DoubleFloats.pi_1o24_d64

DoubleFloats.pi_1o24_t64

DoubleFloats.pi_1o256_d64

DoubleFloats.pi_1o256_t64

DoubleFloats.pi_1o2_d64

DoubleFloats.pi_1o2_t64

DoubleFloats.pi_1o32_d64

DoubleFloats.pi_1o32_t64

DoubleFloats.pi_1o3_d64

DoubleFloats.pi_1o3_t64

DoubleFloats.pi_1o4_d64

DoubleFloats.pi_1o4_t64

DoubleFloats.pi_1o4_t64_hi

DoubleFloats.pi_1o4_t64_lo

DoubleFloats.pi_1o4_t64_md

DoubleFloats.pi_1o512_t64

DoubleFloats.pi_1o64_d64

DoubleFloats.pi_1o64_t64

DoubleFloats.pi_1o6_d64

DoubleFloats.pi_1o6_t64

DoubleFloats.pi_1o8_d64

DoubleFloats.pi_1o8_t64

DoubleFloats.pi_2o1_d64

DoubleFloats.pi_2o1_t64

DoubleFloats.pi_4o1_d64

DoubleFloats.pi_4o1_t64

DoubleFloats.pi_7o16_d64

DoubleFloats.pi_7o16_t64

DoubleFloats.pi_minus

DoubleFloats.posinf

DoubleFloats.powr2_fp_db

DoubleFloats.powr2_fp_dd

DoubleFloats.powr3_fp_db

DoubleFloats.powr3_fp_dd

DoubleFloats.powr4_fp_db

DoubleFloats.powr4_fp_dd

DoubleFloats.powr5_fp_db

DoubleFloats.powr5_fp_dd

DoubleFloats.powr6_fp_db

DoubleFloats.powr6_fp_dd

DoubleFloats.qrtrpi

DoubleFloats.qrtrpi_d64

DoubleFloats.qrtrpi_df64

DoubleFloats.qrtrpi_minus

DoubleFloats.renorm

DoubleFloats.renorm_hilo

DoubleFloats.root2_fp_db

DoubleFloats.root2_fp_dd

DoubleFloats.root3_fp_db

DoubleFloats.root3_fp_dd

DoubleFloats.root4_fp_db

DoubleFloats.root4_fp_dd

DoubleFloats.root5_ddsi_dd

DoubleFloats.root5_fp_db

DoubleFloats.root5_fp_dd

DoubleFloats.root6_fp_db

DoubleFloats.root6_fp_dd

DoubleFloats.rootn_dbsi_db

DoubleFloats.rootn_ddsi_dd

DoubleFloats.sec_pi1o1024_t64

DoubleFloats.sec_pi1o128_t64

DoubleFloats.sec_pi1o12_t64

DoubleFloats.sec_pi1o15_t64

DoubleFloats.sec_pi1o16_t64

DoubleFloats.sec_pi1o2048_t64

DoubleFloats.sec_pi1o24_t64

DoubleFloats.sec_pi1o36_t64

DoubleFloats.sec_pi1o3_t64

DoubleFloats.sec_pi1o48_t64

DoubleFloats.sec_pi1o4_t64

DoubleFloats.sec_pi1o512_t64

DoubleFloats.sec_pi1o5_t64

DoubleFloats.sec_pi1o64_t64

DoubleFloats.sec_pi1o6_t64

DoubleFloats.sec_pi1o7_t64

DoubleFloats.sec_pi1o8_t64

DoubleFloats.sec_taylor

DoubleFloats.showtyped

DoubleFloats.significand

DoubleFloats.signs

DoubleFloats.sin_circle

DoubleFloats.sin_kernel

DoubleFloats.sin_npio32

DoubleFloats.sin_pi1o1024_t64

DoubleFloats.sin_pi1o128_t64

DoubleFloats.sin_pi1o12_t64

DoubleFloats.sin_pi1o15_t64

DoubleFloats.sin_pi1o16_t64

DoubleFloats.sin_pi1o2048_t64

DoubleFloats.sin_pi1o24_t64

DoubleFloats.sin_pi1o36_t64

DoubleFloats.sin_pi1o3_t64

DoubleFloats.sin_pi1o48_t64

DoubleFloats.sin_pi1o4_t64

DoubleFloats.sin_pi1o512_t64

DoubleFloats.sin_pi1o5_t64

DoubleFloats.sin_pi1o64_t64

DoubleFloats.sin_pi1o6_t64

DoubleFloats.sin_pi1o7_t64

DoubleFloats.sin_pi1o8_t64

DoubleFloats.sin_taylor

DoubleFloats.sincos_circle

DoubleFloats.sincos_taylor

DoubleFloats.sixteenthpi

DoubleFloats.sld — Function.

sld(x, y)

Spread the result of x/y.

DoubleFloats.splitnumstring

DoubleFloats.spread — Function.

 spread(x)

the nearest integer to x, away from zero spread complements trunc()

DoubleFloats.sqrt_db_db

DoubleFloats.sqrt_dd_dd

DoubleFloats.square

DoubleFloats.square_db_db

DoubleFloats.square_dd_dd

DoubleFloats.square_fp_dd

DoubleFloats.stringtyped

DoubleFloats.sub2 — Function.

sub2(x::T, y::T) where {T<:AbstractFloat}

Subtract x and y as if they are DoubleFloat{T} numbers. This is more efficient as first converting to DoubleFloat{T} and then subtracting.

DoubleFloats.sub223

DoubleFloats.sub232

DoubleFloats.sub233

DoubleFloats.sub312

DoubleFloats.sub313

DoubleFloats.sub322

DoubleFloats.sub323

DoubleFloats.sub332

DoubleFloats.sub333

DoubleFloats.sub_dbdb_db

DoubleFloats.sub_dbdb_db_nonfinite

DoubleFloats.sub_dbfp_db

DoubleFloats.sub_dbfp_db_nonfinite

DoubleFloats.sub_dddd_dd

DoubleFloats.sub_ddfp_dd

DoubleFloats.sub_fpdb_db

DoubleFloats.sub_fpdb_db_nonfinite

DoubleFloats.sub_fpdd_dd

DoubleFloats.sub_fpfp_db

DoubleFloats.sub_fpfp_dd

DoubleFloats.tan0qrtrpi

DoubleFloats.tan0qrtrpi_denomcoeffs

DoubleFloats.tan0qrtrpi_denompoly

DoubleFloats.tan0qrtrpi_numercoeffs

DoubleFloats.tan0qrtrpi_numerpoly

DoubleFloats.tan_circle

DoubleFloats.tan_coeff

DoubleFloats.tan_npio32

DoubleFloats.tan_pi1o1024_t64

DoubleFloats.tan_pi1o128_t64

DoubleFloats.tan_pi1o12_t64

DoubleFloats.tan_pi1o15_t64

DoubleFloats.tan_pi1o16_t64

DoubleFloats.tan_pi1o2048_t64

DoubleFloats.tan_pi1o24_t64

DoubleFloats.tan_pi1o36_t64

DoubleFloats.tan_pi1o3_t64

DoubleFloats.tan_pi1o48_t64

DoubleFloats.tan_pi1o4_t64

DoubleFloats.tan_pi1o512_t64

DoubleFloats.tan_pi1o5_t64

DoubleFloats.tan_pi1o64_t64

DoubleFloats.tan_pi1o6_t64

DoubleFloats.tan_pi1o7_t64

DoubleFloats.tan_pi1o8_t64

DoubleFloats.tan_taylor

DoubleFloats.thirtysecondpi

DoubleFloats.threesixteenthpi

DoubleFloats.tld — Function.

tld(x, y)

Truncate the result of x/y.

DoubleFloats.triple

DoubleFloats.triple16

DoubleFloats.triple16inv

DoubleFloats.triple32

DoubleFloats.triple32inv

DoubleFloats.triple64

DoubleFloats.triple64inv

DoubleFloats.triple_inv

DoubleFloats.trytypedparse

DoubleFloats.twokminus1

DoubleFloats.twopi

DoubleFloats.twopi_d64

DoubleFloats.twopi_df64

DoubleFloats.twopi_minus

DoubleFloats.ulp

DoubleFloats.value_minus_pi

DoubleFloats.value_minus_qrtrpi

DoubleFloats.⊕

DoubleFloats.⊖

DoubleFloats.⊗

DoubleFloats.⊘