Docstrings

PowerModelsAnnex.ACPPowerModel

PowerModelsAnnex.APIACPPowerModel

PowerModelsAnnex.CostCurve — Type.

abstract type CostCurve end

Abstract type for cost curves. We have two types, one being PolynomialCost, the other PWLCost. These types are introduced in order to deal with the possible formats in which cost curves can be specified and to enforce type (and unit) consistency.

PowerModelsAnnex.DCPPowerModel

PowerModelsAnnex.LOGGER

PowerModelsAnnex.NLACRForm

PowerModelsAnnex.NLACRPowerModel — Type.

default NLACRForm constructor

PowerModelsAnnex.NLForms

PowerModelsAnnex.NLSOCWROAForm

PowerModelsAnnex.NLSOCWRPowerModel — Type.

default NLSOCWROAForm constructor

PowerModelsAnnex.Network — Type.

Network

Basic type for storing a network as a group of DataFrames plus a PowerModels network case.

Constructors:

Network(; bus, gen, piload, psload, line, costgen, costload, pmc)

Create a Network containing the specified dataframes. Specifying a PowerModel case pmc will solely store the case inside the Network, without using it to populate the dataframes. If that is desired, another constructor should be used.

Network(pmc::Dict{String, Any})

Create a Network from a PowerModel network.

Network(path::AbstractString)

Create a Network from a Matpower case file.

Network(case::Symbol)

Create a Network from a case. All cases are stored inside the case_library folder. In order to list all valid case names, use PowerModelsAPI.list_cases().

PowerModelsAnnex.PMs

PowerModelsAnnex.PWLCost — Type.

PWLCost <: CostCurve

The cost function as a piecewise linear function of the MWhr.

Fields

cost::AbstractVector: the vector of the value of the cost function at the extrema

of the segments

mw::AbstractVector: the vector of the extrema of the segments in which

the range for power generation is split.

Constructor

For convenience, and to avoid errors due to the order, the default constructor uses kwargs:

    PWLCost(mw=mw, cost=cost)

NOTE:

Units are explicitly given.

TODO:

Make the dimensions and units parameters of the type.

PowerModelsAnnex.PolynomialCost — Type.

PolynomialCost{T<:Number} <: CostCurve

Field

coefficients::AbstractVector{T}

The coefficients field is a vector that contains the numerical values of the coefficients of the monomials of the polynomial cost function. If n=length(coefficients), the cost is a polynomial of degree n - 1. The first element of the array is the coefficient of the monomial of degree n - 1, the last is the coefficient of the monomial of degree 0.

NOTE:

Dimensions are not enforced as the coefficients have all different dimension (the element i of a vector of length n has dimensions USD/(MWhr)^(n-i))

PowerModelsAnnex.PowerModelsAnnex

PowerModelsAnnex.QCWRPowerModel

PowerModelsAnnex.QCWRTriNoLinkForm

PowerModelsAnnex.QCWRTriNoLinkPowerModel — Type.

default QC trilinear without linking constraint model constructor

PowerModelsAnnex.SOCWROAForm

PowerModelsAnnex.SOCWROAPowerModel — Type.

default SOCWROA constructor

PowerModelsAnnex.SOCWRPowerModel

PowerModelsAnnex._stripunits!

PowerModelsAnnex.add_bus! — Function.

add_bus!(
    net::Network;
    bus_type::Int=0,
    base_kv::Union{Missings.Missing,<:Number}=missing,
    name::AbstractString="none",
    element_id::Int=-1,
    voltage::Union{Missings.Missing,<:Number}=missing,
    volt_max::Union{Missings.Missing,<:Number}=missing,
    volt_min::Union{Missings.Missing,<:Number}=missing,
    coords::Union{Missings.Missing,Tuple{Float64,Float64}}=missing,
)

Add a new bus to Network net. If name and element_id are not provided, reasonable values will be automatically chosen.

PowerModelsAnnex.add_cost_gen! — Function.

add_cost_gen!(
    net::Network;
    coeffs::Vector{<:Real}=Vector{Float64}(),
    gen_id::Union{Missings.Missing,Int}= missing,
    element_id::Int= -1,
    model::Int=2,
    ncost::Int=0,
)

Add new generator cost to a Network net. If element_id is not specified, a reasonable value will be adopted. This method applies to polynomial costs.

add_cost_gen!(
    net::Network;
    pwl_cost::PWLCost;
    gen_id::Union{NAtype,Int}= NA,
    element_id::Int= -1
    model::Int=1,
    ncost::Int=0,
)

Add new generator cost to a Network net. If element_id is not specified, a reasonable value will be adopted. This method applies to piecewise linear costs (1).

PowerModelsAnnex.add_cost_load! — Function.

add_cost_load!(
    net::Network;
    coeffs::Vector{<:Real}=Vector{Float64}(),
    load_id::Union{Missings.Missing,Int}= missing,
    element_id::Int= -1,
    model::Int=2,
    ncost::Int=0,
)

Add new load cost to a Network net. If element_id is not specified, a reasonable value will be adopted. This method applies to polynomial costs.

add_cost_load!(
    net::Network;
    coeffs::Tuple{Vector}=([], []),
    load_id::Union{NAtype,Int}= NA,
    element_id::Int= -1,
    model::Int=1,
    ncost::Int=0,
)

Add new load cost to a Network net. If element_id is not specified, a reasonable value will be adopted. This method applies to piecewise linear costs (1).

PowerModelsAnnex.add_gen! — Function.

add_gen!(
    net::Network;
    element_id::Int=-1,
    bus::Union{Missings.Missing,Int}=missing,
    gen_p::Union{Missings.Missing,<:Number}=missing,
    p_max::Union{Missings.Missing,<:Number}=missing,
    p_min::Union{Missings.Missing,<:Number}=missing,
    gen_q::Union{Missings.Missing,<:Number}=missing,
    q_max::Union{Missings.Missing,<:Number}=missing,
    q_min::Union{Missings.Missing,<:Number}=missing,
    cost::Union{Missings.Missing,Int}=missing,
    startup_cost::Union{Missings.Missing,<:Number}=missing,
    status::Int= 1,
    ramp::Union{Missings.Missing,<:Number}=missing,
)

Add a new generator to a Network net. In case element_id and status are not provided, a reasonable value will be chosen for element_id, while status wil be assumed as 1 (the element is active).

PowerModelsAnnex.add_line! — Function.

add_line!(
    net::Network;
    rate_a::Union{Missings.Missing,<:Number}= missing,
    rate_b::Union{Missings.Missing,<:Number}= missing,
    rate_c::Union{Missings.Missing,<:Number}= missing,
    transformer::Bool= false,
    from_bus::Union{Missings.Missing,Int}= missing,
    to_bus::Union{Missings.Missing,Int}= missing,
    status::Int=1,
    element_id::Int=-1,
    resistance::Union{Missings.Missing,<:Number}= missing,
    reactance::Union{Missings.Missing,<:Number}= missing,
    susceptance::Union{Missings.Missing,<:Number}= missing,
    ang_min::Float64=-1.0,
    ang_max::Float64=1.0,
    p_to::Float64=0.0,
    p_from::Float64=0.0,
    q_to::Float64=0.0,
    q_from::Float64=0.0,
)

Add a new line to a Network net. In case element_id, transformer and status are not provided, a reasonable value will be chosen for element_id, while status wil be assumed as 1 (the element is active) and transformer will be assumed as false (the line is not a transformer).

PowerModelsAnnex.add_load! — Function.

add_load!(
    net::Network;
    element_id::Int=-1,
    bus::Union{Missings.Missing,Int}=missing,
    load_p::Union{Missings.Missing,<:Number}=missing,
    load_q::Union{Missings.Missing,<:Number}=missing,
    status::Int=1,
)

Add a new price insensitive load to a Network net. In case element_id and status are not provided, a reasonable value will be chosen for element_id, while status wil be assumed as 1 (the element is active).

PowerModelsAnnex.add_load_demand_setpoint — Function.

PowerModelsAnnex.add_pi_load! — Function.

add_pi_load!(
    net::Network;
    element_id::Int=-1,
    bus::Union{Missings.Missing,Int}=missing,
    load_p::Union{Missings.Missing,<:Number}=missing,
    load_q::Union{Missings.Missing,<:Number}=missing,
    status::Int=1,
)

Add a new price insensitive load to a Network net. In case element_id and status are not provided, a reasonable value will be chosen for element_id, while status wil be assumed as 1 (the element is active).

PowerModelsAnnex.add_ps_load! — Function.

add_ps_load!(
    net::Network;
    element_id::Int=-1,
    bus::Union{Missings.Missing,Int}=missing,
    load::Union{Missings.Missing,<:Number}=missing,
    load_max::Union{Missings.Missing,<:Number}=missing,
    cost::Union{Missings.Missing,Int}=missing,
    status::Int=1,
)

Add a new price sensitive load to a Network net. In case element_id and status are not provided, a reasonable value will be chosen for element_id, while status wil be assumed as 1 (the element is active).

PowerModelsAnnex.applyunits! — Function.

applyunits!(net::Network)

This function annotates the Network structure with physical unit annotations. For example, for the load column, it is originally in Float64. This function will convert it to a type representing the units u"MWh"). See PowerSystemsUnits.jl for more information. We are assuming energy units as opposed to power units.

PowerModelsAnnex.bounds_tighten_voltage — Function.

PowerModelsAnnex.build_df_from_pmc — Function.

build_df_from_pmc(columns::Dict{Symbol, <: Any}, block::Dict{String,Any})

Return a DataFrame as built from a PowerModels case dictionary. columns specifies which quantites to grab and from where. block corresponds to a dictionary from a PowerModels case.

PowerModelsAnnex.build_pmc! — Function.

build_pmc!(net::Network)

Read information contained in net and uses it to populate net.pmc. This will overwrite any information previously contained in there.

PowerModelsAnnex.bus

PowerModelsAnnex.bus_columns — Constant.

bus_columns

:elementid -> id number of the bus :name -> name of the bus :voltage -> bus operating voltage :voltmax -> maximum bus voltage :voltmin -> minimum bus voltage :bustype -> Bus type :base_kv -> Base voltage :coords -> Bus geocoordinates in latitude and longitude :zone -> Bus zone :area -> Bus Area

PowerModelsAnnex.buses

PowerModelsAnnex.coefficients

PowerModelsAnnex.constraint_kcl_shunt_scaled — Function.

PowerModelsAnnex.constraint_thermal_limit_from — Function.

PowerModelsAnnex.constraint_thermal_limit_to — Function.

PowerModelsAnnex.converged — Function.

converged(net::Network)

Return true if an OPF was ran and successfully converged, false otherwise (inclunding if no OPF was ran).

PowerModelsAnnex.cost

PowerModelsAnnex.cost_columns — Constant.

cost_columns

:element_id -> ID number of the cost (defaults to -1) :coeffs -> coefficients of the cost function (defaults to :cost) :model -> Piecewise linear (1) or polynomial (2), defaults to polynomial (2) :ncost -> Number of coefficients in the polynomial cost or number of pairs (MWh, USD) in piecewise linear costs.

PowerModelsAnnex.cost_gen

PowerModelsAnnex.cost_load

PowerModelsAnnex.costcurve2pmc — Function.

costcurve2pmc(c)

Convert the CostCurve types into PowerModels costs

PowerModelsAnnex.costs

PowerModelsAnnex.degree

PowerModelsAnnex.eval

PowerModelsAnnex.gen

PowerModelsAnnex.gen_columns — Constant.

gen_columns

:elementid -> generator ID number :bus -> ID of the bus where the generator is placed :genp -> active power generated :pmax -> maximum active power :pmin -> minimum active power :genq -> reactive power generated :qmax -> maximum reactive power :qmin -> minimum reactive power :cost -> ID number of the associated cost :startupcost -> startup cost :status -> 0=IDLE, 1=ACTIVE :ramp -> 10-minute ramping constraint

PowerModelsAnnex.gen_cost

PowerModelsAnnex.generator

PowerModelsAnnex.generators

PowerModelsAnnex.gens

PowerModelsAnnex.get_api_solution — Function.

PowerModelsAnnex.include

PowerModelsAnnex.infeasible — Function.

infeasible(net::Network)

Check if the total demand can be satisfied by the generators. Returns true in case no solution is possible.

PowerModelsAnnex.is_convex — Function.

is_convex(mw::AbstractVector{<:Number}, cost::AbstractVector{<:Number})

Determines if the data used for a PWL curve are defining a convex curve.

PowerModelsAnnex.line

PowerModelsAnnex.line_columns — Constant.

line_columns

:elementid -> ID number of the line :ratea -> maximum current under rate A :rateb -> maximum current under rate B :ratec -> maximum current under rate C :transformer -> is transformer? :frombus -> first end-bus ID :tobus -> second end-bus ID :status -> 0=IDLE, 1=ACTIVE :resistance -> line electric resistance :reactance -> line reactance :susceptance -> line susceptance :angmin -> minimum voltage angle difference :angmax -> maximum voltage angle difference :gto -> line conductance (to-node) :gfr -> line conductance (from-node) :bto -> line charging susceptance (to-node) :bfr -> line charging susceptance (from-node)

PowerModelsAnnex.lines

PowerModelsAnnex.lmps — Function.

lmps(net::Network)

Return a DataFrame with the LMPs for all buses after an OPF run. The prices are extracted from the Lagrange multipliers for the Kirchhoff's conservation law. Due to conventions in PowerModels.jl, the LMP are the negative of these Lagrange multipliers.

PowerModelsAnnex.load

PowerModelsAnnex.load_cost

PowerModelsAnnex.loads

PowerModelsAnnex.matpower2pmc

PowerModelsAnnex.max_load_percent! — Function.

max_load_percent!(net::Network, maxload::Real)

Scale line ratings such that the maximum load is equal to maxload% of the original value.

max_load_percent!(pmc::Dict, maxload::Real)

Scale line ratings such that the maximum load is equal to maxload% of the original value.

PowerModelsAnnex.mws

PowerModelsAnnex.n_cost

PowerModelsAnnex.n_segments

PowerModelsAnnex.net2pmc_branch

PowerModelsAnnex.net2pmc_bus

PowerModelsAnnex.net2pmc_gen

PowerModelsAnnex.net2pmc_load

PowerModelsAnnex.net2pmc_ps_load

PowerModelsAnnex.network2pmc — Function.

network2pmc(net::Network)

Return a PowerModels network model (dictionary) based on the information contained within a Network net.

PowerModelsAnnex.objective_max_loading — Function.

objective: Max. load_factor

PowerModelsAnnex.objective_max_loading_gen_output — Function.

PowerModelsAnnex.objective_max_loading_voltage_norm — Function.

PowerModelsAnnex.pi_load

PowerModelsAnnex.pi_load_columns — Constant.

pi_load_columns

:elementid -> ID number of the load :bus -> ID number of the bus where the load is placed :loadp -> active load value :load_q -> reactive load value :status -> 0=IDLE, 1=ACTIVE

PowerModelsAnnex.pmc

PowerModelsAnnex.post_ac_opf — Function.

Given a JuMP model and a PowerModels network data structure, Builds an AC-OPF formulation of the given data and returns the JuMP model

PowerModelsAnnex.post_ac_pf — Function.

Given a JuMP model and a PowerModels network data structure, Builds an AC-PF formulation of the given data and returns the JuMP model

PowerModelsAnnex.post_api_opf — Function.

PowerModelsAnnex.post_dc_opf — Function.

Given a JuMP model and a PowerModels network data structure, Builds an DC-OPF formulation of the given data and returns the JuMP model

PowerModelsAnnex.post_dc_pf — Function.

Given a JuMP model and a PowerModels network data structure, Builds an DC-PF formulation of the given data and returns the JuMP model

PowerModelsAnnex.post_qc_opf — Function.

Given a JuMP model and a PowerModels network data structure, Builds an QC-OPF formulation of the given data and returns the JuMP model Implementation provided by @sidhant172

PowerModelsAnnex.post_sad_opf — Function.

PowerModelsAnnex.post_soc_opf — Function.

Given a JuMP model and a PowerModels network data structure, Builds an SOC-OPF formulation of the given data and returns the JuMP model

PowerModelsAnnex.post_soc_pf — Function.

Given a JuMP model and a PowerModels network data structure, Builds an SOC-PF formulation of the given data and returns the JuMP model

PowerModelsAnnex.prices — Function.

prices(pwl_cost::PWLCost)

This function returns the price for each block of the PWL curve, defined as the slopes of the linear function for each block.

PowerModelsAnnex.ps_load

PowerModelsAnnex.ps_load_columns — Constant.

ps_load_columns

:elementid -> ID number of the load :bus -> ID number of the bus where the load is placed :load -> active load value :loadmax -> maximum active load :cost -> ID number of the associated cost :status -> 0=IDLE, 1=ACTIVE

PowerModelsAnnex.res_gen — Function.

res_gen(net::Network)

Return a DataFrame with the power supplied by each generator as a solution to an OPF.

PowerModelsAnnex.results

PowerModelsAnnex.run_ac_opf

PowerModelsAnnex.run_api_opf — Function.

PowerModelsAnnex.run_dc_opf

PowerModelsAnnex.run_opf

PowerModelsAnnex.run_sad_opf — Function.

PowerModelsAnnex.shadow_prices_lines — Function.

shadow_prices_lines(net::Network)

Return a DataFrame containing the shadow prices for all lines after an OPF run.

PowerModelsAnnex.stripunits!

PowerModelsAnnex.upperbound_negative_active_generation — Function.

PowerModelsAnnex.variable_load_factor — Function.

variable: load_factor >= 1.0