Docstrings

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc. Will always @inline.

julia> using LightQuery

julia> (@_ _ + 1)(1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc. Will always @inline.

julia> using LightQuery

julia> (@_ _ + 1)(1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

LightQuery.By — Type.

By(it, call)

Mark that it has been pre-sorted by call. For use with Group or Join.

julia> using LightQuery

julia> By([1, 2], identity)
By{Array{Int64,1},typeof(identity)}([1, 2], identity)

By(it, call)

Mark that it has been pre-sorted by call. For use with Group or Join.

julia> using LightQuery

julia> By([1, 2], identity)
By{Array{Int64,1},typeof(identity)}([1, 2], identity)

LightQuery.CSV

LightQuery.File

LightQuery.Filter

LightQuery.Generator

LightQuery.Group — Type.

Group(it::By)

Group consecutive keys in it. Requires a presorted object (see By). Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types.

julia> using LightQuery

julia> Group(By([1, 3, 2, 4], iseven)) |> collect
2-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 0 => [1, 3]
 1 => [2, 4]

Group(it::By)

Group consecutive keys in it. Requires a presorted object (see By). Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types.

julia> using LightQuery

julia> Group(By([1, 3, 2, 4], iseven)) |> collect
2-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 0 => [1, 3]
 1 => [2, 4]

LightQuery.History

LightQuery.IteratorLength

LightQuery.Join — Type.

Join(left::By, right::By)

Find all pairs where isequal(left.call(left.it), right.call(right.it)).

julia> using LightQuery

julia> Join(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        collect
6-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
       1 => 1
       2 => missing
 missing => 3
 missing => 4
       5 => missing
       6 => 6

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first.

julia> @> [1, 1, 2, 2] |>
        Group(By(_, identity)) |>
        By(_, first) |>
        Join(_, By([1, 2], identity)) |>
        collect
2-element Array{Pair{Pair{Int64,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},Int64},1}:
 (1 => [1, 1]) => 1
 (2 => [2, 2]) => 2

For other join flavors, combine with when. Make sure to annotate with Length if you know it.

julia> @> Join(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        when(_, @_ !ismissing(_.first)) |>
        Length(_, 4) |>
        collect
4-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

Join(left::By, right::By)

Find all pairs where isequal(left.call(left.it), right.call(right.it)).

julia> using LightQuery

julia> Join(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        collect
6-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
       1 => 1
       2 => missing
 missing => 3
 missing => 4
       5 => missing
       6 => 6

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first.

julia> @> [1, 1, 2, 2] |>
        Group(By(_, identity)) |>
        By(_, first) |>
        Join(_, By([1, 2], identity)) |>
        collect
2-element Array{Pair{Pair{Int64,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},Int64},1}:
 (1 => [1, 1]) => 1
 (2 => [2, 2]) => 2

For other join flavors, combine with when. Make sure to annotate with Length if you know it.

julia> @> Join(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        when(_, @_ !ismissing(_.first)) |>
        Length(_, 4) |>
        collect
4-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

LightQuery.Length — Type.

Length(it, length)

Allow optimizations based on length. Especially useful after Join and before make_columns.

julia> using LightQuery

julia> @> Filter(iseven, 1:4) |>
        Length(_, 2) |>
        collect
2-element Array{Int64,1}:
 2
 4

Length(it, length)

Allow optimizations based on length. Especially useful after Join and before make_columns.

julia> using LightQuery

julia> @> Filter(iseven, 1:4) |>
        Length(_, 2) |>
        collect
2-element Array{Int64,1}:
 2
 4

LightQuery.LightQuery

LightQuery.Name — Type.

Name(name)

Create a typed name. Can be used as a function to getproperty, with a default to missing. For multiple names, see Names.

julia> using LightQuery

julia> (a = 1,) |>
        Name(:a)
1

julia> (a = 1,) |>
        Name(:b)
missing

julia> missing |>
        Name(:a)
missing

Name(name)

Create a typed name. Can be used as a function to getproperty, with a default to missing. For multiple names, see Names.

julia> using LightQuery

julia> (a = 1,) |>
        Name(:a)
1

julia> (a = 1,) |>
        Name(:b)
missing

julia> missing |>
        Name(:a)
missing

LightQuery.Names — Type.

Names(the_names...)

Create typed names. Can be used to as a function to assign or select names, with a default to missing. For just one name, see Name.

julia> using LightQuery

julia> (1, 1.0) |>
        Names(:a, :b)
(a = 1, b = 1.0)

julia> (a = 1, b = 1.0) |>
        Names(:a)
(a = 1,)

julia> (a = 1,) |>
        Names(:a, :b)
(a = 1, b = missing)

julia> missing |>
        Names(:a)
(a = missing,)

Names(the_names...)

Create typed names. Can be used to as a function to assign or select names, with a default to missing. For just one name, see Name.

julia> using LightQuery

julia> (1, 1.0) |>
        Names(:a, :b)
(a = 1, b = 1.0)

julia> (a = 1, b = 1.0) |>
        Names(:a)
(a = 1,)

julia> (a = 1,) |>
        Names(:a, :b)
(a = 1, b = missing)

julia> missing |>
        Names(:a)
(a = missing,)

LightQuery.Peek — Type.

Peek(it; max_rows = 4)

Get a peek of an iterator which returns named tuples. Will show no more than max_rows. Relies on item_names.

julia> using LightQuery

julia> (a = 1:5, b = 5:-1:1) |>
        rows |>
        Peek
Showing 4 of 5 rows
|  :a |  :b |
| ---:| ---:|
|   1 |   5 |
|   2 |   4 |
|   3 |   3 |
|   4 |   2 |

Peek(it; max_rows = 4)

Get a peek of an iterator which returns named tuples. Will show no more than max_rows. Relies on item_names.

julia> using LightQuery

julia> (a = 1:5, b = 5:-1:1) |>
        rows |>
        Peek
Showing 4 of 5 rows
|  :a |  :b |
| ---:| ---:|
|   1 |   5 |
|   2 |   4 |
|   3 |   3 |
|   4 |   2 |

LightQuery.ZippedArrays

LightQuery.anonymous

LightQuery.chain

LightQuery.columns — Function.

columns(it)

Inverse of rows. Always lazy, see make_columns for eager version.

julia> using LightQuery

julia> (a = [1], b = [1.0]) |>
        rows |>
        columns
(a = [1], b = [1.0])

columns(it)

Inverse of rows. Always lazy, see make_columns for eager version.

julia> using LightQuery

julia> (a = [1], b = [1.0]) |>
        rows |>
        columns
(a = [1], b = [1.0])

LightQuery.combine_iterator_eltype

LightQuery.eval

LightQuery.first_fallback

LightQuery.flatten

LightQuery.full_dispatch

LightQuery.gather — Function.

gather(it; assignments...)

For each key => value pair in assignments, gather the Names in value into a single key. Inverse of spread.

julia> using LightQuery

julia> @> (a = 1, b = 1.0, c = 1//1) |>
        gather(_, d = Names(:a, :c))
(b = 1.0, d = (a = 1, c = 1//1))

gather(it; assignments...)

For each key => value pair in assignments, gather the Names in value into a single key. Inverse of spread.

julia> using LightQuery

julia> @> (a = 1, b = 1.0, c = 1//1) |>
        gather(_, d = Names(:a, :c))
(b = 1.0, d = (a = 1, c = 1//1))

LightQuery.getproperty_default

LightQuery.include

LightQuery.inner_name

LightQuery.item_names — Function.

item_names(it)

Find names of items in it. Used in Peek and make_columns.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        item_names
(:a, :b)

If inference cannot detect names, it will use propertynames of the first item. Map Names over it to override this behavior.

julia> [(a = 1,), (a = 2, b = 2.0)] |>
        Peek
|  :a |
| ---:|
|   1 |
|   2 |

julia> @> [(a = 1,), (a = 2, b = 2.0)] |>
        over(_, Names(:a, :b)) |>
        Peek
|  :a |      :b |
| ---:| -------:|
|   1 | missing |
|   2 |     2.0 |

item_names(it)

Find names of items in it. Used in Peek and make_columns.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        item_names
(:a, :b)

If inference cannot detect names, it will use propertynames of the first item. Map Names over it to override this behavior.

julia> [(a = 1,), (a = 2, b = 2.0)] |>
        Peek
|  :a |
| ---:|
|   1 |
|   2 |

julia> @> [(a = 1,), (a = 2, b = 2.0)] |>
        over(_, Names(:a, :b)) |>
        Peek
|  :a |      :b |
| ---:| -------:|
|   1 | missing |
|   2 |     2.0 |

LightQuery.item_names_dispatch

LightQuery.key — Function.

key(it)

The key in a key => value pair.

key(it)

The key in a key => value pair.

LightQuery.make_columns — Function.

make_columns(it)

Collect into columns. Always eager, see columns for lazy version. Relies on item_names.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        make_columns
(a = [1, 2], b = [1.0, 2.0])

make_columns(it)

Collect into columns. Always eager, see columns for lazy version. Relies on item_names.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        make_columns
(a = [1, 2], b = [1.0, 2.0])

LightQuery.maybe_push_widen

LightQuery.maybe_setindex_widen_up_to

LightQuery.named_tuple — Function.

named_tuple(it)

Coerce to a named_tuple. For performance with working with arbitrary structs, define and @inline propertynames.

julia> using LightQuery

julia> @inline Base.propertynames(p::Pair) = (:first, :second)

julia> named_tuple(:a => 1)
(first = :a, second = 1)

named_tuple(it)

Coerce to a named_tuple. For performance with working with arbitrary structs, define and @inline propertynames.

julia> using LightQuery

julia> @inline Base.propertynames(p::Pair) = (:first, :second)

julia> named_tuple(:a => 1)
(first = :a, second = 1)

LightQuery.next_history

LightQuery.order — Function.

order(it, call; keywords...)

Generalized sort. keywords will be passed to sort!; see the documentation there for options. See By for a way to explicitly mark that an object has been sorted.

julia> using LightQuery

julia> order([2, 1], identity)
2-element view(::Array{Int64,1}, [2, 1]) with eltype Int64:
 1
 2

    order(it, call, condition; keywords...)

If `call` is not type stable, consider adding a `condition` to filter.

jldoctest order julia> using LightQuery

julia> order([2, 1, missing], identity, !ismissing) 2-element view(::Array{Union{Missing, Int64},1}, [2, 1]) with eltype Union{Missing, Int64}: 1 2 ```

LightQuery.over — Function.

over(it, call)

Lazy map with argument order reversed.

over(it, call)

Lazy map with argument order reversed.

LightQuery.remove — Function.

remove(it, the_names...)

Remove the_names. Inverse of transform.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        remove(_, :b)
(a = 1,)

remove(it, the_names...)

Remove the_names. Inverse of transform.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        remove(_, :b)
(a = 1,)

LightQuery.rename — Function.

rename(it; renames...)

Rename it. Because constants do not constant propagate through key-word arguments, wrap with Name.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        rename(_, c = Name(:a))
(b = 1.0, c = 1)

rename(it; renames...)

Rename it. Because constants do not constant propagate through key-word arguments, wrap with Name.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        rename(_, c = Name(:a))
(b = 1.0, c = 1)

LightQuery.rows — Function.

rows(it)

Iterator over rows of a NamedTuple of arrays. Always lazy. Inverse of columns. See Peek for a way to view.

julia> using LightQuery

julia> (a = [1, 2], b = [1.0, 2.0]) |>
        rows |>
        collect
2-element Array{NamedTuple{(:a, :b),Tuple{Int64,Float64}},1}:
 (a = 1, b = 1.0)
 (a = 2, b = 2.0)

rows(it)

Iterator over rows of a NamedTuple of arrays. Always lazy. Inverse of columns. See Peek for a way to view.

julia> using LightQuery

julia> (a = [1, 2], b = [1.0, 2.0]) |>
        rows |>
        collect
2-element Array{NamedTuple{(:a, :b),Tuple{Int64,Float64}},1}:
 (a = 1, b = 1.0)
 (a = 2, b = 2.0)

LightQuery.spread — Function.

spread(it::NamedTuple, the_names...)

Unnest nested it in name. Inverse of gather.

julia> using LightQuery

julia> @> (b = 1.0, d = (a = 1, c = 1//1)) |>
        spread(_, :d)
(b = 1.0, a = 1, c = 1//1)

spread(it::NamedTuple, the_names...)

Unnest nested it in name. Inverse of gather.

julia> using LightQuery

julia> @> (b = 1.0, d = (a = 1, c = 1//1)) |>
        spread(_, :d)
(b = 1.0, a = 1, c = 1//1)

LightQuery.state_to_index

LightQuery.substitute_underscores!

LightQuery.transform — Function.

transform(it; assignments...)

Merge assignments into it. Inverse of remove.

julia> using LightQuery

julia> @> (a = 1,) |>
        transform(_, b = 1.0)
(a = 1, b = 1.0)

transform(it; assignments...)

Merge assignments into it. Inverse of remove.

julia> using LightQuery

julia> @> (a = 1,) |>
        transform(_, b = 1.0)
(a = 1, b = 1.0)

LightQuery.unzip — Function.

unzip(it, n)

Unzip an iterator it which returns tuples of length n. Use Val(n) to guarantee type stability.

julia> using LightQuery

julia> unzip([(1, 1.0), (2, 2.0)], 2)
([1, 2], [1.0, 2.0])

julia> unzip([(1, 1.0), (2, 2.0)], Val(2))
([1, 2], [1.0, 2.0])

unzip(it, n)

Unzip an iterator it which returns tuples of length n. Use Val(n) to guarantee type stability.

julia> using LightQuery

julia> unzip([(1, 1.0), (2, 2.0)], 2)
([1, 2], [1.0, 2.0])

julia> unzip([(1, 1.0), (2, 2.0)], Val(2))
([1, 2], [1.0, 2.0])

LightQuery.value — Function.

value(it)

The value in a key => value pair.

value(it)

The value in a key => value pair.

LightQuery.when — Function.

when(it, call)

Lazy filter with argument order reversed.

when(it, call)

Lazy filter with argument order reversed.