Docstrings

SortingLab.Bits104

SortingLab.Bits112

SortingLab.Bits120

SortingLab.Bits24

SortingLab.Bits40

SortingLab.Bits48

SortingLab.Bits56

SortingLab.Bits72

SortingLab.Bits80

SortingLab.Bits88

SortingLab.Bits96

SortingLab.SortingLab

SortingLab.VT

SortingLab.VTV

SortingLab._fsortperm_msd_hybrid

SortingLab.eval

SortingLab.fsort — Function.

radixsort(svec, rev = false)

Applies radix sort to the string vector, svec, and sort it in place.

SortingLab.fsort!

SortingLab.fsort2

SortingLab.fsort2!

SortingLab.fsortandperm_int_range_lsd

SortingLab.fsortperm — Function.

fsortperm(svec)

Faster sortperm for string vectors

SortingLab.fsortperm_int_range_counting_sort

SortingLab.fsortperm_int_range_lsd

SortingLab.fsortperm_msd_hybrid

SortingLab.include

SortingLab.load_bits

SortingLab.load_bits_with_padding

SortingLab.load_uint

SortingLab.radixsort

SortingLab.radixsort! — Function.

radixsort!(svec)

Applies radix sort to the string vector, svec, and sort it in place.

SortingLab.sort_spointer!

SortingLab.sorttwo! — Function.

sorttwo!(vs, index)

Sort both the vs and reorder index at the same. This allows for faster sortperm for radix sort.

SortingLab.uint_hist — Function.

uint_hist(bits, [RADIX_SIZE = 16, RADIX_MASK = 0xffff])

Computes a histogram (counts) for the vector RADIXSIZE bits at a time. E.g. if eltype(bits) is UInt64 and RADIXSIZE is 16 then 4 histograms are created for each of the 16 bit chunks.