Docstrings

BigCombinatorics.Bell — Function.

Bell(n) gives the n-th Bell number, that is, the number of partitions of an n-element set.

BigCombinatorics.BigCombinatorics

BigCombinatorics.Binomial — Function.

Binomial(n,k) returns the binomial coefficient n-choose-k. This is the number of k-element subsets of an n-element set.

BigCombinatorics.Catalan — Function.

Catalan(n) returns the n-th Catalan number.

BigCombinatorics.Derangements — Function.

Derangements(n) returns the number of permutations of an n-set that have no fixed point.

BigCombinatorics.DoubleFactorial — Function.

DoubleFactorial(n) returns n!!, i.e., n*(n-2)*... with (-1)!! == 0!! == 1!! == 1.

BigCombinatorics.Euler — Function.

Euler(n) returns the n-th Euler number. Starting with n=0 this is the sequence 1, 0, -1, 0, 5, 0, -61, 0, 1385 and so on.

BigCombinatorics.Eulerian — Function.

Eulerian(n,k) returns the number of permutations of {1,2,...,n} with k ascents.

BigCombinatorics.Factorial — Function.

Factorial(n) returns n! for nonnegative integers n. Factorial(n,k) returns n!/k! (to be consistent with Julia's factorial.) Requires 0 <= k <= n.

See also FallingFactorial and RisingFactorial.

BigCombinatorics.FallingFactorial — Function.

FallingFactorial(n,k) returns n*(n-1)*(n-2)*...*(n-k+1) (with a total of k factors). Requires n,k >= 0. If k>n then 0 is returned.

BigCombinatorics.Fibonacci — Function.

Fibonacci(n) returns the n-th Fibonacci number. We begin with Fibonacci(0)==0 and Fibonacci(1)==1.

BigCombinatorics.IntPartitions — Function.

IntPartitions(n) is the number of partitions of the integer n.

IntPartitions(n,k) is the number of partitions of the integer n with exactly k (nonzero) parts.

BigCombinatorics.IntPartitionsDistinct — Function.

IntPartitionsDistinct(n,k) is the number of partitions of the integer n into exactly k distinct parts.

IntPartitionsDistinct(n) is the number of partitions of n into distinct parts.

BigCombinatorics.MultiChoose — Function.

MultiChoose(n,k) returns the number of k-element multisets that can be formed using the elements of an n-element set.

Warning: This is not the same as Multinomial.

BigCombinatorics.Multinomial — Function.

Multinomial(vec) returns the multinomial coefficient whose top index is the sum of vec (an array of Ints) and whose bottom indices are given by vec.

This may also be called with a common-separated list of arguments, that is, either of Multinomial([1,2,3]) or Multinomial(1,2,3). The result is 60 in both cases as these equal 6!/(1! 2! 3!).

Warning: This is not the same as MultiChoose.

BigCombinatorics.PowerSum — Function.

PowerSum(n,k) returns the sum of the k-th powers of the integers 1 through n, i.e., 1^k + 2^k + 3^k + ... + n^k.

BigCombinatorics.RisingFactorial — Function.

RisingFactorial(n,k) returns n*(n+1)*(n+2)*...*(n+k-1) (with a total of k factors). Requires n,k >= 0.

BigCombinatorics.Stirling1 — Function.

Stirling1(n,k) gives the (signed) Stirling number of the first kind, that is, the coefficient of x^k in the poynomial x(x-1)(x-2)...(x-n+1).

BigCombinatorics.Stirling2 — Function.

Stirling2(n,k) gives the Stirling number of the second kind, that is, the number of paritions of an n-set into k-parts."

BigCombinatorics._get

BigCombinatorics._has

BigCombinatorics._make

BigCombinatorics._master_table

BigCombinatorics._save

BigCombinatorics.cache_clear — Function.

BigCombinatorics.cache_clear(f) clears the cached values for the function f. Returns false if f values are not cached.

BigCombinatorics.cache_clear() clears all cached values.

BigCombinatorics.cache_report — Function.

BigCombinatorics.cache_report() reports how many entries are saved for each function in the BigCombinatorics module.

BigCombinatorics.eval

BigCombinatorics.include