Docstrings

MPIFiles./

MPIFiles.@B

MPIFiles.@Np

MPIFiles.@affineunit

MPIFiles.@cNp

MPIFiles.@dB

MPIFiles.@dateformat_str

MPIFiles.@derived_dimension

MPIFiles.@dimension

MPIFiles.@logscale

MPIFiles.@logunit

MPIFiles.@refunit

MPIFiles.@u_str

MPIFiles.@unit

MPIFiles.ARPACKException

MPIFiles.AbstractFFTs

MPIFiles.AbstractRNG

MPIFiles.Adjoint

MPIFiles.Apr

MPIFiles.April

MPIFiles.ArbitraryPositions

MPIFiles.Aug

MPIFiles.August

MPIFiles.Axis

MPIFiles.AxisAlignedBox

MPIFiles.AxisArray

MPIFiles.BLAS

MPIFiles.Ball

MPIFiles.Bidiagonal

MPIFiles.BreakpointGridPositions

MPIFiles.BrukerDatasetStore

MPIFiles.BrukerFile

MPIFiles.BrukerFileCalib

MPIFiles.BrukerFileFast

MPIFiles.BrukerFileMeas

MPIFiles.BrukerStore

MPIFiles.BunchKaufman

MPIFiles.ChebyshevGridPositions

MPIFiles.Cholesky

MPIFiles.CholeskyPivoted

MPIFiles.ColorView

MPIFiles.ColoringParams

MPIFiles.DatasetStore

MPIFiles.Date

MPIFiles.DateFormat

MPIFiles.DatePeriod

MPIFiles.DateTime

MPIFiles.Dates

MPIFiles.Day

MPIFiles.Dec

MPIFiles.December

MPIFiles.DelimitedFiles

MPIFiles.Diagonal

MPIFiles.DimensionlessQuantity

MPIFiles.Eigen

MPIFiles.Experiment

MPIFiles.FFTW

MPIFiles.Factorization

MPIFiles.Feb

MPIFiles.February

MPIFiles.Fri

MPIFiles.Friday

MPIFiles.Gain

MPIFiles.GeneralizedEigen

MPIFiles.GeneralizedSVD

MPIFiles.GeneralizedSchur

MPIFiles.GridPositions

MPIFiles.HTSS

MPIFiles.HasTimeAxis

MPIFiles.Hermitian

MPIFiles.Hessenberg

MPIFiles.Hour

MPIFiles.I

MPIFiles.IMTFile

MPIFiles.IMTFileCalib

MPIFiles.IMTFileMeas

MPIFiles.ISODateFormat

MPIFiles.ISODateTimeFormat

MPIFiles.ISOTimeFormat

MPIFiles.ImageAxes

MPIFiles.ImageCore

MPIFiles.ImageMeta

MPIFiles.ImageMetadata

MPIFiles.IndexAny

MPIFiles.IndexIncremental

MPIFiles.JCVAL

MPIFiles.Jan

MPIFiles.January

MPIFiles.JcampdxFile

MPIFiles.Jul

MPIFiles.July

MPIFiles.Jun

MPIFiles.June

MPIFiles.LAPACK

MPIFiles.LAPACKException

MPIFiles.LDLt

MPIFiles.LQ

MPIFiles.LU

MPIFiles.Level

MPIFiles.LinearAlgebra

MPIFiles.LowerTriangular

MPIFiles.MDFDatasetStore

MPIFiles.MDFFile

MPIFiles.MDFFileV1

MPIFiles.MDFFileV2

MPIFiles.MDFStore

MPIFiles.MPIFile

MPIFiles.MPIFiles

MPIFiles.Mar

MPIFiles.March

MPIFiles.May

MPIFiles.MeanderingGridPositions

MPIFiles.MersenneTwister

MPIFiles.Microsecond

MPIFiles.Millisecond

MPIFiles.Minute

MPIFiles.MixingOrder

MPIFiles.Mmap

MPIFiles.Mon

MPIFiles.Monday

MPIFiles.Month

MPIFiles.MultiMPIFile

MPIFiles.MultiPatchMPIFile

MPIFiles.Nanosecond

MPIFiles.NoDims

MPIFiles.NoUnits

MPIFiles.Nov

MPIFiles.November

MPIFiles.Oct

MPIFiles.October

MPIFiles.Period

MPIFiles.PosDefException

MPIFiles.Positions

MPIFiles.QR

MPIFiles.QRPivoted

MPIFiles.Quantity

MPIFiles.RFC1123Format

MPIFiles.Random

MPIFiles.RandomDevice

MPIFiles.RankDeficientException

MPIFiles.Reconstruction

MPIFiles.RegularGridPositions

MPIFiles.SVD

MPIFiles.Sat

MPIFiles.Saturday

MPIFiles.Schur

MPIFiles.Second

MPIFiles.Sep

MPIFiles.September

MPIFiles.SingularException

MPIFiles.SpatialDomain

MPIFiles.SphericalTDesign

MPIFiles.StackedView

MPIFiles.Statistics

MPIFiles.StreamIndexStyle

MPIFiles.StreamingContainer

MPIFiles.Study

MPIFiles.Sun

MPIFiles.Sunday

MPIFiles.SymTridiagonal

MPIFiles.Symmetric

MPIFiles.Thu

MPIFiles.Thursday

MPIFiles.Time

MPIFiles.TimeAxis

MPIFiles.TimePeriod

MPIFiles.TimeType

MPIFiles.TimeZone

MPIFiles.Transpose

MPIFiles.Tridiagonal

MPIFiles.Tue

MPIFiles.Tuesday

MPIFiles.UTC

MPIFiles.UniformRandomPositions

MPIFiles.UniformScaling

MPIFiles.UnitLowerTriangular

MPIFiles.UnitUpperTriangular

MPIFiles.Unitful

MPIFiles.UpperTriangular

MPIFiles.Visualization

MPIFiles.Wed

MPIFiles.Wednesday

MPIFiles.Week

MPIFiles.Year

MPIFiles.\

MPIFiles._innerGenerateSFDatabase

MPIFiles._interpolate_inner

MPIFiles._iscalib

MPIFiles._makeStringArray

MPIFiles._mixingFactors

MPIFiles.absoluteunit

MPIFiles.acqFov

MPIFiles.acqFramePeriod

MPIFiles.acqGradient

MPIFiles.acqGradientDiag

MPIFiles.acqNumAverages

MPIFiles.acqNumBGFrames

MPIFiles.acqNumFGFrames

MPIFiles.acqNumFrames

MPIFiles.acqNumPatches

MPIFiles.acqNumPeriods

MPIFiles.acqNumPeriodsPerFrame

MPIFiles.acqNumPeriodsPerPatch

MPIFiles.acqOffsetField

MPIFiles.acqOffsetFieldShift

MPIFiles.acqSize

MPIFiles.acqStartTime

MPIFiles.activeChannels

MPIFiles.addLeadingSingleton

MPIFiles.addReco

MPIFiles.addStudy

MPIFiles.addTrailingSingleton

MPIFiles.addVisu

MPIFiles.adjoint

MPIFiles.adjoint!

MPIFiles.adjust

MPIFiles.appendBGDataset

MPIFiles.assert_timedim_last

MPIFiles.atindex

MPIFiles.atvalue

MPIFiles.axes

MPIFiles.axisdim

MPIFiles.axisnames

MPIFiles.axisvalues

MPIFiles.axpby!

MPIFiles.axpy!

MPIFiles.ballRadius

MPIFiles.bfft

MPIFiles.bfft!

MPIFiles.bitrand

MPIFiles.brfft

MPIFiles.bunchkaufman

MPIFiles.bunchkaufman!

MPIFiles.calcPrefactors

MPIFiles.calculateSystemMatrixSNR

MPIFiles.calibDeltaSampleSize

MPIFiles.calibFov

MPIFiles.calibFovCenter

MPIFiles.calibIsMeanderingGrid

MPIFiles.calibMethod

MPIFiles.calibOffsetField

MPIFiles.calibOrder

MPIFiles.calibPositions

MPIFiles.calibSNR

MPIFiles.calibSize

MPIFiles.calibdir

MPIFiles.canonicalize

MPIFiles.channelview

MPIFiles.cheb2D

MPIFiles.cholesky

MPIFiles.cholesky!

MPIFiles.clamp01

MPIFiles.clamp01nan

MPIFiles.collapse

MPIFiles.colordim

MPIFiles.colorsigned

MPIFiles.colorview

MPIFiles.cond

MPIFiles.condskeel

MPIFiles.convertr

MPIFiles.convertrp

MPIFiles.converttometer

MPIFiles.converttoreal

MPIFiles.coords_spatial

MPIFiles.copy_transpose!

MPIFiles.copyproperties

MPIFiles.copyto!

MPIFiles.cor

MPIFiles.cov

MPIFiles.cross

MPIFiles.currentSetting

MPIFiles.data

MPIFiles.datetime2julian

MPIFiles.datetime2rata

MPIFiles.datetime2unix

MPIFiles.day

MPIFiles.dayabbr

MPIFiles.dayname

MPIFiles.dayofmonth

MPIFiles.dayofquarter

MPIFiles.dayofweek

MPIFiles.dayofweekofmonth

MPIFiles.dayofyear

MPIFiles.daysinmonth

MPIFiles.daysinyear

MPIFiles.daysofweekinmonth

MPIFiles.dct

MPIFiles.dct!

MPIFiles.defaultParams

MPIFiles.deltaSampleConcentration

MPIFiles.deltaSampleVolume

MPIFiles.deriveSubgrid

MPIFiles.det

MPIFiles.dfBaseFrequency

MPIFiles.dfCustomWaveform

MPIFiles.dfCycle

MPIFiles.dfDivider

MPIFiles.dfNumChannels

MPIFiles.dfPhase

MPIFiles.dfStrength

MPIFiles.dfWaveform

MPIFiles.diag

MPIFiles.diagind

MPIFiles.diagm

MPIFiles.dimension

MPIFiles.done_

MPIFiles.dot

MPIFiles.eigen

MPIFiles.eigen!

MPIFiles.eigmax

MPIFiles.eigmin

MPIFiles.eigvals

MPIFiles.eigvals!

MPIFiles.eigvecs

MPIFiles.eval

MPIFiles.experimentDescription

MPIFiles.experimentHasMeasurement

MPIFiles.experimentHasProcessing

MPIFiles.experimentHasReconstruction

MPIFiles.experimentIsCalibration

MPIFiles.experimentIsSimulation

MPIFiles.experimentName

MPIFiles.experimentNumber

MPIFiles.experimentSubject

MPIFiles.experimentUuid

MPIFiles.exportToMDFStore

MPIFiles.factorize

MPIFiles.fft

MPIFiles.fft!

MPIFiles.fftshift

MPIFiles.fieldOfView

MPIFiles.fieldOfViewCenter

MPIFiles.filepath

MPIFiles.filterFrequencies

MPIFiles.findBrukerFiles

MPIFiles.findReco

MPIFiles.findSFFiles

MPIFiles.findfirst_

MPIFiles.firstdayofmonth

MPIFiles.firstdayofquarter

MPIFiles.firstdayofweek

MPIFiles.firstdayofyear

MPIFiles.float32

MPIFiles.float64

MPIFiles.fullFramePermutation

MPIFiles.generateSFDatabase

MPIFiles.generateSFDatabase_

MPIFiles.getAveragedMeasurements

MPIFiles.getComplexType

MPIFiles.getExperiment

MPIFiles.getExperiments

MPIFiles.getInterpolator

MPIFiles.getMeasurements

MPIFiles.getMeasurementsFD

MPIFiles.getMeasurementsLowLevel

MPIFiles.getNewCalibNum

MPIFiles.getNewExperimentNum

MPIFiles.getNewNumInFolder

MPIFiles.getPermutation

MPIFiles.getReco

MPIFiles.getRecons

MPIFiles.getStudies

MPIFiles.getStudy

MPIFiles.getSystemMatrix

MPIFiles.getSystemMatrixReshaped

MPIFiles.getVisu

MPIFiles.getVisuPath

MPIFiles.getVisus

MPIFiles.getindex!

MPIFiles.getmask

MPIFiles.givens

MPIFiles.h5exists

MPIFiles.hannWindow

MPIFiles.hasKeyAndValue

MPIFiles.height

MPIFiles.hessenberg

MPIFiles.hessenberg!

MPIFiles.hour

MPIFiles.id

MPIFiles.idct

MPIFiles.idct!

MPIFiles.idxToPos

MPIFiles.ifft

MPIFiles.ifft!

MPIFiles.ifftshift

MPIFiles.imcenter

MPIFiles.include

MPIFiles.indexPermutation

MPIFiles.indices_spatial

MPIFiles.irfft

MPIFiles.isComplexArray

MPIFiles.isSubgrid

MPIFiles.isdiag

MPIFiles.ishermitian

MPIFiles.ishidden

MPIFiles.isleapyear

MPIFiles.isposdef

MPIFiles.isposdef!

MPIFiles.issuccess

MPIFiles.issymmetric

MPIFiles.istimeaxis

MPIFiles.istril

MPIFiles.istriu

MPIFiles.julian2datetime

MPIFiles.kron

MPIFiles.lastdayofmonth

MPIFiles.lastdayofquarter

MPIFiles.lastdayofweek

MPIFiles.lastdayofyear

MPIFiles.latin1toutf8

MPIFiles.ldiv!

MPIFiles.ldlt

MPIFiles.ldlt!

MPIFiles.linear

MPIFiles.lmul!

MPIFiles.loadDataset

MPIFiles.loadMetadata

MPIFiles.loadMetadataOnline

MPIFiles.loadParams

MPIFiles.loadRecoDataMDF

MPIFiles.loadRecoDataMDF_

MPIFiles.loadSFDatabase

MPIFiles.loadTDesign — Function.

Returns the t-Design Array for choosen t and N.

MPIFiles.logabsdet

MPIFiles.logdet

MPIFiles.logunit

MPIFiles.lowrankdowndate

MPIFiles.lowrankdowndate!

MPIFiles.lowrankupdate

MPIFiles.lowrankupdate!

MPIFiles.lq

MPIFiles.lq!

MPIFiles.lu

MPIFiles.lu!

MPIFiles.lyap

MPIFiles.mean

MPIFiles.mean!

MPIFiles.meanderingFramePermutation

MPIFiles.measBGFrameBlockLengths

MPIFiles.measBGFrameIdx

MPIFiles.measData

MPIFiles.measDataConv

MPIFiles.measDataLowLevel

MPIFiles.measDataSpectralLeakageCorrected

MPIFiles.measDataSpectralLeakageCorrectedMultiPatch

MPIFiles.measDataSpectralLeakageCorrectedSinglePatch

MPIFiles.measDataTD

MPIFiles.measDataTDPeriods

MPIFiles.measFGFrameIdx

MPIFiles.measFramePermutation

MPIFiles.measIsBGCorrected

MPIFiles.measIsBGFrame

MPIFiles.measIsFourierTransformed

MPIFiles.measIsFramePermutation

MPIFiles.measIsFrequencySelection

MPIFiles.measIsFrequencySelecton

MPIFiles.measIsSpectralLeakageCorrected

MPIFiles.measIsTFCorrected

MPIFiles.measIsTransposed

MPIFiles.median

MPIFiles.median!

MPIFiles.microsecond

MPIFiles.middle

MPIFiles.millisecond

MPIFiles.minute

MPIFiles.mixFactorToFreq — Function.

This function returns the multiplicator k corresponding to the frequency given by the mixing factors mx, my, and mz. I.e. the integer value, for which kF = mxfx + myfy + mz*fz, where F is the measurement cycle frequency and fx, fy, and fz are the excitation frequencies for the x,y, and z channel.

MPIFiles.mixFactorToFreqIdx — Function.

This function returns the index freqidx of the frequency given by the mixing factors mx, my, and mz with respect to the frequency list freq = frequencies(bSF).

MPIFiles.mixingFactors — Function.

This function returns a lookup table with columns freqidx | mx | my | mz | mixingOrder for all frequencies in freq = frequencies(bSF), where only the lowest order mixing coefficients mx, my, and mz are listed.

MPIFiles.month

MPIFiles.monthabbr

MPIFiles.monthday

MPIFiles.monthname

MPIFiles.mul!

MPIFiles.n0f16

MPIFiles.n0f8

MPIFiles.n2f14

MPIFiles.n4f12

MPIFiles.n6f10

MPIFiles.nanosecond

MPIFiles.next_

MPIFiles.nimages

MPIFiles.norm

MPIFiles.normalize

MPIFiles.normalize!

MPIFiles.normedview

MPIFiles.now

MPIFiles.nullspace

MPIFiles.numCurrentSettings

MPIFiles.numLatitude

MPIFiles.numMeridian

MPIFiles.numSubPeriods — Function.

This function is used as a workaround that no 1D and 2D sequences were functional. Therefore we acquired 3D data and set the corresponding drive-field amplitude to zero. Stepsize is then the number of period that fit into the full 3D sequence.

MPIFiles.opnorm

MPIFiles.ordschur

MPIFiles.ordschur!

MPIFiles.paddedviews

MPIFiles.permuteddimsview

MPIFiles.pinv

MPIFiles.pixelspacing

MPIFiles.plan_bfft

MPIFiles.plan_bfft!

MPIFiles.plan_brfft

MPIFiles.plan_dct

MPIFiles.plan_dct!

MPIFiles.plan_fft

MPIFiles.plan_fft!

MPIFiles.plan_idct

MPIFiles.plan_idct!

MPIFiles.plan_ifft

MPIFiles.plan_ifft!

MPIFiles.plan_irfft

MPIFiles.plan_rfft

MPIFiles.posToIdx

MPIFiles.posToIdxFloat

MPIFiles.posToLinIdx

MPIFiles.properties

MPIFiles.qr

MPIFiles.qr!

MPIFiles.quantile

MPIFiles.quantile!

MPIFiles.quarterofyear

MPIFiles.radius

MPIFiles.rand!

MPIFiles.randcycle

MPIFiles.randcycle!

MPIFiles.randexp

MPIFiles.randexp!

MPIFiles.randn!

MPIFiles.randperm

MPIFiles.randperm!

MPIFiles.randstring

MPIFiles.randsubseq

MPIFiles.randsubseq!

MPIFiles.rank

MPIFiles.rata2datetime

MPIFiles.rawview

MPIFiles.rdiv!

MPIFiles.readComplexArray

MPIFiles.readdlm

MPIFiles.readonly

MPIFiles.recoData

MPIFiles.recoFov

MPIFiles.recoFovCenter

MPIFiles.recoOrder

MPIFiles.recoParameters

MPIFiles.recoPositions

MPIFiles.recoSize

MPIFiles.reflevel

MPIFiles.reinterpretc

MPIFiles.remove

MPIFiles.returnasreal

MPIFiles.rfft

MPIFiles.rmul!

MPIFiles.rxBandwidth

MPIFiles.rxDataConversionFactor

MPIFiles.rxFrequencies

MPIFiles.rxInductionFactor

MPIFiles.rxNumChannels

MPIFiles.rxNumFrequencies

MPIFiles.rxNumSamplingPoints

MPIFiles.rxTimePoints

MPIFiles.rxTransferFunction

MPIFiles.rxUnit

MPIFiles.saveParams

MPIFiles.saveRecoDataMDF

MPIFiles.saveasMDF

MPIFiles.saveasMDFHacking

MPIFiles.scaleminmax

MPIFiles.scalesigned

MPIFiles.scannerFacility

MPIFiles.scannerManufacturer

MPIFiles.scannerName

MPIFiles.scannerOperator

MPIFiles.scannerTopology

MPIFiles.schur

MPIFiles.schur!

MPIFiles.sdims

MPIFiles.second

MPIFiles.seed

MPIFiles.selectedChannels

MPIFiles.selectedReceivers

MPIFiles.setparam!

MPIFiles.sfPath

MPIFiles.shape

MPIFiles.shareproperties

MPIFiles.shuffle

MPIFiles.shuffle!

MPIFiles.size_spatial

MPIFiles.spacedirections

MPIFiles.spacing

MPIFiles.spatialorder

MPIFiles.spatialproperties

MPIFiles.start_

MPIFiles.std

MPIFiles.stdm

MPIFiles.str2uuid

MPIFiles.studyDescription

MPIFiles.studyName

MPIFiles.studyNumber

MPIFiles.studyUuid

MPIFiles.studydir

MPIFiles.svd

MPIFiles.svd!

MPIFiles.svdvals

MPIFiles.svdvals!

MPIFiles.sylvester

MPIFiles.systemMatrix

MPIFiles.systemMatrixWithBG

MPIFiles.takemap

MPIFiles.timeaxis

MPIFiles.timedim

MPIFiles.today

MPIFiles.tofirst

MPIFiles.tolast

MPIFiles.tonext

MPIFiles.toprev

MPIFiles.tr

MPIFiles.tracerBatch

MPIFiles.tracerConcentration

MPIFiles.tracerInjectionTime

MPIFiles.tracerName

MPIFiles.tracerSolute

MPIFiles.tracerVendor

MPIFiles.tracerVolume

MPIFiles.transpose

MPIFiles.transpose!

MPIFiles.tril

MPIFiles.tril!

MPIFiles.triu

MPIFiles.triu!

MPIFiles.try_chmod

MPIFiles.uconvert

MPIFiles.uconvertp

MPIFiles.uconvertrp

MPIFiles.unflattenOffsetFieldShift

MPIFiles.unit

MPIFiles.unix2datetime

MPIFiles.upreferred

MPIFiles.ustrip

MPIFiles.uuid

MPIFiles.var

MPIFiles.varm

MPIFiles.version

MPIFiles.week

MPIFiles.width

MPIFiles.widthheight

MPIFiles.writeComplexArray

MPIFiles.writeIfAvailable

MPIFiles.writedlm

MPIFiles.year

MPIFiles.yearmonth

MPIFiles.yearmonthday

MPIFiles.zeroarray

MPIFiles.×

MPIFiles.⋅