Docstrings

ReusePatterns.@copy_fields — Macro.

@copy_fields T

Copy all field definitions from a structure into another.

Example

julia> struct First
    field1
    field2::Int
end
julia> struct Second
    @copy_fields(First)
    field3::String
end

julia> println(fieldnames(Second))
(:field1, :filed2, :field3)

julia> println(fieldtype.(Second, fieldnames(Second)))
(Any, Int64, String)

ReusePatterns.@forward — Macro.

@forward(sender, receiver, ekws...)

Evaluate the Julia code to forward methods. The syntax is exactly the same as the forward function.

Example:

julia> struct Book
    title::String
    author::String
end
julia> Base.show(io::IO, b::Book) = println(io, "$(b.title) (by $(b.author))")
julia> Base.print(b::Book) = println("In a hole in the ground there lived a hobbit...")
julia> author(b::Book) = b.author

julia> struct PaperBook
    b::Book
    number_of_pages::Int
end
julia> @forward((PaperBook, :b), Book)
julia> pages(book::PaperBook) = book.number_of_pages

julia> struct Edition
    b::PaperBook
    year::Int
end
julia> @forward((Edition, :b), PaperBook)
julia> year(book::Edition) = book.year

julia> book = Edition(PaperBook(Book("The Hobbit", "J.R.R. Tolkien"), 374), 2013)

julia> print(author(book), ", ", pages(book), " pages, Ed. ", year(book))
J.R.R. Tolkien, 374 pages, Ed. 2013

julia> print(book)
In a hole in the ground there lived a hobbit...

ReusePatterns.@quasiabstract — Macro.

@quasiabstract expression

Create a quasi abstract type.

This macro accepts an expression defining a (mutable or immutable) structure, and outputs the code for two new type definitions:

an abstract type with the same name and (if given) supertype of the input structure;

a concrete structure definition with name prefix Concrete_, subtyping the abstract type defined above.

The relation between the types ensure there will be a single concrete type associated to the abstract one, hence you can use the abstract type to annotate method arguments, and be sure to receive the associated concrete type, or one of its subtypes (which shares all field names and types of the ancestors).

The concrete type associated to an quasi abstract type can be retrieved with the concretetype function. The Concrete_ prefix can be customized passing a second symbol argument to the macro.

These newly defined types allows to easily implement concrete subtyping: simply use the quasi abstract type name for both object construction and to annotate method arguments.

Example:

julia> @quasiabstract struct Book
    title::String
    author::String
end
julia> Base.show(io::IO, b::Book) = println(io, "$(b.title) (by $(b.author))")
julia> Base.print(b::Book) = println("In a hole in the ground there lived a hobbit...")
julia> author(b::Book) = b.author

julia> @quasiabstract struct PaperBook <: Book
    number_of_pages::Int
end
julia> pages(book::PaperBook) = book.number_of_pages

julia> @quasiabstract struct Edition <: PaperBook
    year::Int
end
julia> year(book::Edition) = book.year

julia> println(fieldnames(concretetype(Edition)))
(:title, :author, :number_of_pages, :year)

julia> book = Edition("The Hobbit", "J.R.R. Tolkien", 374, 2013)

julia> print(author(book), ", ", pages(book), " pages, Ed. ", year(book))
J.R.R. Tolkien, 374 pages, Ed. 2013

julia> print(book)
In a hole in the ground there lived a hobbit...

julia> isquasiconcrete(typeof(book))
true

julia> isquasiabstract(supertype(typeof(book)))
true

julia> concretetype(typeof(book)) === supertype(typeof(book))
true

ReusePatterns.ReusePatterns

ReusePatterns.concretetype — Function.

concretetype(T::Type)

Return the concrete type associated with the quasi abstract type T. If T is not quasi abstract returns nothing.

See also: @quasiabstract

ReusePatterns.eval

ReusePatterns.forward — Function.

forward(sender::Tuple{Type,Symbol}, receiver::Type, method::Method; withtypes=true, allargs=true)

Return a Vector{String} containing the Julia code to properly forward method calls to from a sender type to a receiver type.

The sender tuple must contain a structure type, and a symbol with the name of one of its fields.

The withtypes keyword controls whether the forwarding method has type annotated arguments. The allargs keyword controls wether all arguments should be used, or just the first ones up to the last containing the receiver type.

Both keywords are true by defult, but they can be set to false to decrease the number of forwarding methods.

Example:

Implement a wrapper for an Int object, and forward the + method accepting Int:

struct Wrapper{T}
    wrappee::T
    extra
    Wrapper{T}(args...; kw...) where T = new(T(args...; kw...), nothing)
end

# Prepare and evaluate forwarding methods:
m = forward((Wrapper, :wrappee), Int, which(+, (Int, Int)))
eval.(Meta.parse.(m))

# Instantiate two wrapped `Int`
i1 = Wrapper{Int}(1)
i2 = Wrapper{Int}(2)

# And add them seamlessly
println(i1 +  2)
println( 1 + i2)
println(i1 + i2)

forward(sender::Tuple{Type,Symbol}, receiver::Type, methods::Vector{Method}; kw...)

forward(sender::Tuple{Type,Symbol}, receivers::Vector{T}, methods; kw...)

forward(sender::Tuple{Type,Symbol}, receiver::Type; super=true, kw...)

Wrapper for forward(send, receiver, supertypes(receiver, super=super); kw...)

ReusePatterns.include

ReusePatterns.isquasiabstract — Function.

isquasiabstract(T::Type)

Return true if T is quasi abstract, false otherwise.

See also: @quasiabstract

ReusePatterns.isquasiconcrete — Function.

isquasiconcrete(T::Type)

Return true if T is a concrete type associated to a quasi abstract type, false otherwise.

See also: @quasiabstract

ReusePatterns.supertypes — Function.

supertypes(T::Type)

Returns a vector with all supertypes of type T (excluding Any).

Example

julia> println(supertypes(Int))
Type[Signed, Integer, Real, Number]