Docstrings

MarkovChains.ContMarkovChain

MarkovChains.CtmcSolution

MarkovChains.FintimeCumSolution

MarkovChains.FintimeProbSolution

MarkovChains.FintimeSolution

MarkovChains.Graphs

MarkovChains.InftimeStateResult

MarkovChains.Layouts

MarkovChains.MarkovChains

MarkovChains.Reorder

MarkovChains.Transition

MarkovChains.aa_rate_matrix — Function.

aa_rate_matrix(chain, order, abs_start, nabs)

create transition rate matrix (to-from order) from between recurrent states within range absstart, absstart + abs

MarkovChains.add_state!

MarkovChains.add_transition!

MarkovChains.chain_to_lightgraph

MarkovChains.eval

MarkovChains.extend!

MarkovChains.fintime_solve

MarkovChains.fintime_solve_cum

MarkovChains.fintime_solve_prob

MarkovChains.include

MarkovChains.inftime_solve — Function.

inftime_state(chain, init_prob; spsolve)

compute state cumulative times/probabilites of the markov chain at time infinity. state_prob and state_cumtime can be used to retrieve times/probs from the return value.

MarkovChains.max_out_rate

MarkovChains.mean_time_to_absorption

MarkovChains.plot_chain

MarkovChains.poisson_cum_rtp

MarkovChains.poisson_trunc_point

MarkovChains.reorder_states — Function.

reorder states into transient states, recurrent components.

MarkovChains.solve

MarkovChains.state_count

MarkovChains.state_cumtime — Function.

state_cumtime(result, state)

used to retrieve state cumulative time from result.

MarkovChains.state_prob — Function.

state_prob(result, state)

used to retrieve state probablity from result.

MarkovChains.states

MarkovChains.ta_rate_matrix — Function.

ta_rate_matrix(chain, order, abs_start, nabs)

create transition rate matrix (to-from order) from transient states to recurrent states within range absstart, absstart + abs

MarkovChains.trans_prob_matrix

MarkovChains.trans_rate_matrix

MarkovChains.transitions

MarkovChains.tt_rate_matrix — Function.

tt_rate_matrix(chain, mat2chain, chain2mat, ntrans)

create transition rate matrix (to-from order) between transient states.