Docstrings

PeriodicTable.Element — Type.

Element composite type

PeriodicTable.Elements — Type.

Elements composite type

PeriodicTable.PeriodicTable — Module.

The PeriodicTable module exports an elements variable that stores data (of type Element) on every element in the periodic table.

The data can be looked up by atomic number via elements[n], by name (case-insensitive) via e.g. elements["oxygen"], or by symbol via e.g. elements[:O].

PeriodicTable._elements_data

PeriodicTable.elements

PeriodicTable.eval

PeriodicTable.include

PeriodicTable.ispresent

PeriodicTable.printpresent

PeriodicTable.printpresenthtml