Docstrings

StaticNumbers.@staticnumbers — Macro.

@staticnumbers numbers 1argfuns 2argfuns

This macro creates methods that return Static numbers when functions are called with only Static arguments.

The inputs should be a list of literal numbers that will be tested as inputs and outputs of all functions.

Optionally a fourth argument can give a list of numbers that will only be considered as results, but not as inputs.

Example:

@staticnumbers (0, 1) (sin, cos) (+, -)

will create all the following method definitions:

sin(::StaticInteger{0}) = StaticInteger{0}()
cos(::StaticInteger{0}) = StaticInteger{1}()
+(::StaticInteger{0}, ::StaticInteger{0}) = StaticInteger{0}()
+(::StaticInteger{0}, ::StaticInteger{1}) = StaticInteger{1}()
+(::StaticInteger{1}, ::StaticInteger{0}) = StaticInteger{1}()
-(::StaticInteger{0}, ::StaticInteger{0}) = StaticInteger{0}()
-(::StaticInteger{1}, ::StaticInteger{0}) = StaticInteger{0}()
-(::StaticInteger{1}, ::StaticInteger{1}) = StaticInteger{1}()

(Note: The macro will run in the local scope. Functions from Base must be imported before they can be extended.)

StaticNumbers.LengthRange — Constant.

LengthRange(zeroth, step, length)

A LengthRange is a range which allows (but does not require) its zeroth, step and/or length to be Static numbers.

The zeroth element of the range is the element before the first. Although it is not part of the range, it is frequently useful for it to remain static. For example, if the common 1:n range is multiplied by a scalar, the zeroth element can remain StaticInteger{0}

The step is the distance between successive elements of the range.

The length must be an Integer. A LengthRange is parameterized by its length, rather than its last element. This makes it possible for the length to remain Static when an offset is added to the range.

LengthRange is the union of LengthStepRange and LengthUnitRange which are subtypes of OrdinalRange and AbstractUnitRange respectively.

StaticNumbers.LengthRangeError

StaticNumbers.LengthStepRange — Type.

A LengthStepRange is a range which is parameterized by its zeroth element, step and length (as oppsed to first, step, and last). The type of each of the parameters is part of the range type, and hence it is possible to use this type of range efficiently with Static parameters.

StaticNumbers.LengthUnitRange — Type.

A LengthUnitRange is a type that is identical to LengthStepRange but where the step is fixed to 1. It is a subtype of UnitRange.

(It would be much simpler to just use LengthStepRange with a Static step of 1 but then methods that expect a UnitRange would not work.)

StaticNumbers.Static — Constant.

Static{X} is short-hand for the Union of StaticInteger{X}, StaticReal{X} and StaticNumber{X}.

StaticNumbers.StaticBool

StaticNumbers.StaticError

StaticNumbers.StaticInteger — Type.

A StaticInteger is an Integer whose value is stored in the type, and which contains no runtime data.

StaticNumbers.StaticNumber — Type.

A StaticNumber is a Number whose value is stored in the type, and which contains no runtime data.

StaticNumbers.StaticNumbers

StaticNumbers.StaticOneTo — Type.

StaticOneTo{N} - Like Base.OneTo{Int}(N) but with the length fixed by the type.

StaticNumbers.StaticOrBool

StaticNumbers.StaticOrInt

StaticNumbers.StaticReal — Type.

A StaticReal is a Real whose value is stored in the type, and which contains no runtime data.

StaticNumbers.eval

StaticNumbers.fma

StaticNumbers.genstaticmethod1 — Function.

Create an expression a method for Static input to fun if the result is among the targets

StaticNumbers.genstaticmethod2 — Function.

Create an expression a method for two Static inputs to fun if the result is among the targets

StaticNumbers.genstaticmethods1 — Function.

Create an array of all the methods for Static 1-argument inputs to fun where the result is either among the possible inputs, args, or among the additional targets.

StaticNumbers.genstaticmethods2 — Function.

Create an array of all the methods for Static 2-argument inputs to fun where the result is either among the possible inputs, args, or among the additional targets.

StaticNumbers.include

StaticNumbers.ofstatictype — Function.

ofstatictype(x,y) - like oftype(x,y), but return a Static x is a Static.

StaticNumbers.static — Function.

static(X) is shorthand for StaticInteger{X}(), StaticReal{X}() or StaticNumber{X}(), depending on the type of X.

StaticNumbers.staticlength — Function.

staticlength(range) converts to a range where the length is Static.

StaticNumbers.staticmod — Function.

staticmod(x,y) returns static(mod(x,y)) if y is Static.

If y is not Static, then staticmod(x,y) is the same as mod(x,y).

StaticNumbers.tostatic — Function.

tostatic(x, r) Returns a Static integer, equal to x from the range r. If no element in r is equal to x, then the behaviour of this function is undefined.

StaticNumbers.tostaticexpr

StaticNumbers.trystatic — Function.

trystatic(x, y1, y2, ...) x ⩢ y1 ⩢ y2 ...

Test if a number x is equal to any of the numbers y1, y2, ..., and in that case return static(y). Otherwise, or if x is already a Static number, `x is returned unchanged.

The inferred return type will typically be a small Union, which Julia can handle efficiently.

This function can be used to call specialized methods for certain input values. For example, f(x, y ⩢ 0) will call f(x, y) if y is nonzero, but f(x, static(0)) if y is zero. This is useful if it enables optimizations that outweigh the cost of branching.

NOTE: When the list of y-values is longer than one, y1, y2, ... must be Static numbers, or inferrence will not work. (In which case trystatic is not more efficient than static(x).)

trystatic(x, r) Tests if an integer x is in the range r, and if so, returns a Static integer from the range. (Otherwise, x is returned unchanged.)

NOTE: The range must be completely static, or inferrence will not work.

StaticNumbers.zeroth

StaticNumbers.⩢