Docstrings

UMAP.SMOOTH_K_TOLERANCE

UMAP.UMAP

UMAP.UMAP_

UMAP.combine_fuzzy_sets

UMAP.compute_membership_strengths — Function.

compute_membership_strengths(knns, dists, σs, ρs) -> rows, cols, vals

Compute the membership strengths for the 1-skeleton of each fuzzy simplicial set.

UMAP.eval

UMAP.fit_ϕ — Function.

fit_ϕ(min_dist, spread) -> a, b

Find a smooth approximation to the membership function of points embedded in ℜᵈ. This fits a smooth curve that approximates an exponential decay offset by min_dist.

UMAP.fuzzy_set_intersection

UMAP.fuzzy_set_union

UMAP.fuzzy_simplicial_set — Function.

fuzzy_simplicial_set(X, n_neighbors, metric, local_connectivity, set_op_ratio) -> graph::SparseMatrixCSC

Construct the local fuzzy simplicial sets of each point in X by finding the approximate nearest n_neighbors, normalizing the distances on the manifolds, and converting the metric space to a simplicial set.

UMAP.include

UMAP.initialize_embedding

UMAP.knn_search — Function.

knn_search(X, k, metric) -> knns, dists

Find the k nearest neighbors of each point in X by metric.

UMAP.optimize_embedding — Function.

optimize_embedding(graph, embedding, n_epochs, initial_alpha, min_dist, spread, gamma, neg_sample_rate) -> embedding

Optimize an embedding by minimizing the fuzzy set cross entropy between the high and low dimensional simplicial sets using stochastic gradient descent.

Arguments

graph: a sparse matrix of shape (nsamples, nsamples)

embedding: a dense matrix of shape (ncomponents, nsamples)

n_epochs: the number of training epochs for optimization

initial_alpha: the initial learning rate

gamma: the repulsive strength of negative samples

neg_sample_rate::Integer: the number of negative samples per positive sample

UMAP.smooth_knn_dist

UMAP.smooth_knn_dists — Function.

smooth_knn_dists(dists, k, local_connectivity; <kwargs>) -> knn_dists, nn_dists

Compute the distances to the nearest neighbors for a continuous value k. Returns the approximated distances to the kth nearest neighbor (knn_dists) and the nearest neighbor (nn_dists) from each point.

UMAP.spectral_layout — Function.

spectral_layout(graph, embed_dim) -> embedding

Initialize the graph layout with spectral embedding.

UMAP.umap — Function.

umap(X::AbstractMatrix[, n_components=2]; <kwargs>) -> embedding

Embed the data X into a n_components-dimensional space. n_neighbors controls how many neighbors to consider as locally connected.

Keyword Arguments

n_neighbors::Integer = 15: the number of neighbors to consider as locally connected. Larger values capture more global structure in the data, while small values capture more local structure.

metric::SemiMetric = Euclidean(): the metric to calculate distance in the input space

n_epochs::Integer = 300: the number of training epochs for embedding optimization

learning_rate::AbstractFloat = 1.: the initial learning rate during optimization

init::Symbol = :spectral: how to initialize the output embedding; valid options are :spectral and :random

min_dist::AbstractFloat = 0.1: the minimum spacing of points in the output embedding

spread::AbstractFloat = 1.0: the effective scale of embedded points. Determines how clustered embedded points are in combination with min_dist.

set_operation_ratio::AbstractFloat = 1.0: interpolates between fuzzy set union and fuzzy set intersection when constructing the UMAP graph (global fuzzy simplicial set). The value of this parameter should be between 1.0 and 0.0: 1.0 indicates pure fuzzy union, while 0.0 indicates pure fuzzy intersection.

local_connectivity::Integer = 1: the number of nearest neighbors that should be assumed to be locally connected. The higher this value, the more connected the manifold becomes. This should not be set higher than the intrinsic dimension of the manifold.

repulsion_strength::AbstractFloat = 1.0: the weighting of negative samples during the optimization process.

neg_sample_rate::Integer = 5: the number of negative samples to select for each positive sample. Higher values will increase computational cost but result in slightly more accuracy.

a::AbstractFloat = nothing: this controls the embedding. By default, this is determined automatically by min_dist and spread.

b::AbstractFloat = nothing: this controls the embedding. By default, this is determined automatically by min_dist and spread.