Docstrings

JLD2.@load

JLD2.@lookup_committed

JLD2.@read_datatype

JLD2.@save

JLD2.AttributeHeader

JLD2.BOXED_READ_DATASPACE

JLD2.BasicDatatype

JLD2.BufferedReader

JLD2.BufferedWriter

JLD2.CHECKSUM_POS

JLD2.CSET_ASCII

JLD2.CSET_UTF8

JLD2.CURRENT_VERSION

JLD2.CharacterSet

JLD2.CommittedDatatype

JLD2.CompactStorageMessage

JLD2.CompoundDatatype

JLD2.ContiguousStorageMessage

JLD2.CustomSerialization — Type.

CustomSerialization{T,S}

On-disk representation for data that is written as if it were of Julia type T, but is read as type S.

JLD2.DEFAULT_BUFFER_SIZE

JLD2.DEFLATE_PIPELINE_MESSAGE

JLD2.DS_NULL

JLD2.DS_SCALAR

JLD2.DS_SIMPLE

JLD2.DT_ARRAY

JLD2.DT_BITFIELD

JLD2.DT_COMPOUND

JLD2.DT_ENUMERATED

JLD2.DT_FIXED_POINT

JLD2.DT_FLOATING_POINT

JLD2.DT_OPAQUE

JLD2.DT_REFERENCE

JLD2.DT_STRING

JLD2.DT_TIME

JLD2.DT_VARIABLE_LENGTH

JLD2.DataMode

JLD2.DataTypeODR

JLD2.DataspaceStart

JLD2.EMPTY_DIMENSIONS

JLD2.EMPTY_READ_ATTRIBUTES

JLD2.FILE_GROW_SIZE

JLD2.FILE_HEADER

JLD2.FILE_HEADER_LENGTH

JLD2.FixedLengthString

JLD2.FixedPointDatatype

JLD2.FloatTypes

JLD2.FloatingPointDatatype

JLD2.GLOBAL_HEAP_SIGNATURE

JLD2.GlobalHeap — Type.

GlobalHeap

Represents an HDF5 global heap structure.

JLD2.GlobalHeapID

JLD2.Group — Type.

Group{T}

JLD group object.

JLD2.H5Datatype — Type.

H5Datatype

Supertype of all HDF5 datatypes.

JLD2.H5TYPE_DATATYPE

JLD2.H5TYPE_UNION

JLD2.H5TYPE_VLEN_UTF8

JLD2.HM_ATTRIBUTE

JLD2.HM_ATTRIBUTE_INFO

JLD2.HM_BOGUS

JLD2.HM_BTREE_K_VALUES

JLD2.HM_DATASPACE

JLD2.HM_DATATYPE

JLD2.HM_DATA_LAYOUT

JLD2.HM_DRIVER_INFO

JLD2.HM_EXTERNAL_FILE_LIST

JLD2.HM_FILL_VALUE

JLD2.HM_FILL_VALUE_OLD

JLD2.HM_FILTER_PIPELINE

JLD2.HM_GROUP_INFO

JLD2.HM_LINK_INFO

JLD2.HM_LINK_MESSAGE

JLD2.HM_MODIFICATION_TIME

JLD2.HM_NIL

JLD2.HM_OBJECT_COMMENT

JLD2.HM_OBJECT_HEADER_CONTINUATION

JLD2.HM_REFERENCE_COUNT

JLD2.HM_SHARED_MESSAGE_TABLE

JLD2.HM_SYMBOL_TABLE

JLD2.HasReferences

JLD2.HeaderMessage

JLD2.IndirectPointer — Type.

IndirectPointer

When writing data, we may need to enlarge the memory mapping, which would invalidate any memory addresses arising from the old mmap pointer. IndirectPointer holds a pointer to the startptr field of an MmapIO, and the offset relative to that pointer. It defers computing a memory address until converted to a Ptr{T}, so the memory mapping can be enlarged and addresses will remain valid.

JLD2.Initialized

JLD2.InternalError

JLD2.InvalidDataException

JLD2.JLD2

JLD2.JLDFile — Type.

JLDFile{T<:IO}

JLD file object.

JLD2.JLDWriteSession — Type.

JLDWriteSession{T}

A JLDWriteSession keeps track of references to serialized objects. If T is a Dict, h5offset maps an object ID (returned by calling objectid) to th RelOffset of the written dataset. If it is Union{}, then references are not tracked, and objects referenced multiple times are written multiple times.

JLD2.LC_CHUNKED_STORAGE

JLD2.LC_COMPACT_STORAGE

JLD2.LC_CONTIGUOUS_STORAGE

JLD2.LM_CREATION_ORDER_PRESENT

JLD2.LM_LINK_NAME_CHARACTER_SET_FIELD_PRESENT

JLD2.LM_LINK_TYPE_FIELD_PRESENT

JLD2.Length

JLD2.LinkInfo

JLD2.Lookup3

JLD2.MMAP_CUTOFF

JLD2.MMAP_GROW_SIZE

JLD2.MmapIO

JLD2.NCHECKSUM

JLD2.NULL_COMMITTED_DATATYPE

JLD2.NULL_REFERENCE

JLD2.OBJECT_HEADER_CONTINUATION_SIGNATURE

JLD2.OBJECT_HEADER_SIGNATURE

JLD2.OH_ATTRIBUTE_CREATION_ORDER_INDEXED

JLD2.OH_ATTRIBUTE_CREATION_ORDER_TRACKED

JLD2.OH_ATTRIBUTE_PHASE_CHANGE_VALUES_STORED

JLD2.OH_TIMES_STORED

JLD2.OPEN_FILES

JLD2.ObjectStart

JLD2.OnDiskRepresentation

JLD2.OpaqueDatatype

JLD2.Plain

JLD2.PlainType

JLD2.PointerException

JLD2.Pointers

JLD2.PrimitiveTypeTypes

JLD2.PrimitiveTypes

JLD2.REQUIRED_FILE_HEADER

JLD2.ReadAttribute

JLD2.ReadDataspace

JLD2.ReadRepresentation — Type.

ReadRepresentation{T,ODR}

A type encoding both the Julia type T and the on-disk (HDF5) representation ODR.

JLD2.ReconstructedTypes

JLD2.ReferenceDatatype

JLD2.ReferenceFree

JLD2.RelOffset — Type.

RelOffset

Represents an HDF5 relative offset. This differs from a file offset (used elsewhere) in that it is relative to the superblock base address. In practice, this means that FILE_HEADER_LENGTHhas been subtracted. fileoffset and h5offset convert between RelOffsets and file offsets.

JLD2.SUPERBLOCK_SIGNATURE

JLD2.SignedTypes

JLD2.StringDatatype

JLD2.Superblock

JLD2.TypeMappingException — Type.

constructrr(f::JLDFile, T::DataType, dt::CompoundType, attrs::Vector{ReadAttribute}, hard_failure::Bool=false)

Constructs a ReadRepresentation for a given type. This is the generic method for all types not specially handled below.

If hard_failure is true, then throw a TypeMappingException instead of attempting reconstruction. This helps in cases where we can't know if reconstructed parametric types will have a matching memory layout without first inspecting the memory layout.

JLD2.UNDEFINED_ADDRESS

JLD2.UndefinedFieldException

JLD2.UnknownType

JLD2.UnsignedTypes

JLD2.UnsupportedFeatureException

JLD2.UnsupportedVersionException

JLD2.VariableLengthDatatype

JLD2.Vlen

JLD2.WriteDataspace

JLD2.WrittenAttribute

JLD2._odr

JLD2._read

JLD2._write

JLD2.adjust_position!

JLD2.begin_checksum_read

JLD2.begin_checksum_write

JLD2.behead — Function.

behead(T)

Given a UnionAll type, recursively eliminates the where clauses

JLD2.check_empty

JLD2.check_writtenas_type

JLD2.chunked_storage_message_size

JLD2.class

JLD2.commit

JLD2.commit_compound

JLD2.construct_array — Function.

construct_array{T}(io::IO, ::Type{T}, ndims::Int)

Construct array by reading ndims dimensions from io. Assumes io has already been seeked to the correct position.

JLD2.constructrr

JLD2.datamode

JLD2.define_packed

JLD2.deflate_data

JLD2.end_checksum

JLD2.ensureroom

JLD2.eval

JLD2.fieldnames

JLD2.fieldodr

JLD2.fileoffset — Function.

fileoffset(f::JLDFile, x::RelOffset)

Converts an offset x relative to the superblock of file f to an absolute offset.

JLD2.find_dimensions_attr

JLD2.finish!

JLD2.get_ndims_offset

JLD2.group_continuation_size

JLD2.group_payload_size — Function.

group_payload_size(pairs)

Returns the size of a group payload, including link info, group info, and link messages, but not the object header. pairs is an iterator of String => RelOffset pairs. Provides space after the last object message for a continuation message.

JLD2.grow

JLD2.h5convert!

JLD2.h5convert_uninitialized!

JLD2.h5fieldtype

JLD2.h5offset — Function.

h5offset(f::JLDFile, x::RelOffset)

Converts an absolute file offset x to an offset relative to the superblock of file f.

JLD2.h5type

JLD2.hasdata

JLD2.hasfielddata

JLD2.heap_object_length

JLD2.include

JLD2.isatend

JLD2.isgroup

JLD2.ismutabletype

JLD2.jlconvert

JLD2.jlconvert_canbeuninitialized

JLD2.jlconvert_isinitialized

JLD2.jld_finalizer — Function.

jld_finalizer(f::JLDFile)

When a JLDFile is finalized, it is possible that the MmapIO has been munmapped, since Julia does not guarantee finalizer order. This means that the underlying file may be closed before we get a chance to write to it.

JLD2.jldopen — Function.

jldopen(fname::AbstractString, mode::AbstractString)

Opens a JLD file at path fname.

"r": Open for reading only, failing if no file exists "r+": Open for reading and writing, failing if no file exists "w"/"w+": Open for reading and writing, overwriting the file if it already exists "a"/"a+": Open for reading and writing, creating a new file if none exists, but preserving the existing file if one is present

JLD2.jltype

JLD2.link_size — Function.

link_size(name::String)

Returns the size of a link message, including message header.

JLD2.links_size — Function.

links_size(pairs)

Returns the size of several link messages. pairs is an iterator of String => RelOffset pairs.

JLD2.load

JLD2.load_dataset

JLD2.load_datatypes — Function.

load_datatypes(f::JLDFile)

Populate f.datatypes and f.jlh5types with all of the committed datatypes from a file. We need to do this before writing to make sure we reuse written datatypes.

JLD2.load_group

JLD2.loadtodict!

JLD2.lookup_offset — Function.

lookup_offset(g::Group, name::AbstractString) -> RelOffset

Lookup the offset of a dataset in a group. Returns UNDEFINED_ADDRESS if the dataset is not present. Does not inspect unwritten_child_groups.

JLD2.mmap!

JLD2.msync

JLD2.munmap

JLD2.numel

JLD2.objodr

JLD2.odr

JLD2.odr_sizeof

JLD2.openfile

JLD2.pathize — Function.

pathize(g::Group, name::AbstractString, create::Bool) -> Tuple{Group,String}

Converts a path to a group and name object. If create is true, any intermediate groups will be created, and the dataset name will be checked for uniqueness with existing names.

JLD2.payload_size_without_storage_message

JLD2.prewrite — Function.

prewrite(f::JLDFile)

Check that a JLD file is actually writable, and throw an error if not. Sets the written flag on the file.

JLD2.raw_write

JLD2.rconvert

JLD2.read_array

JLD2.read_array! — Function.

read_array!(v::Array, f::JLDFile, rr)

Fill the array v with the contents of JLDFile f at the current position, assuming a ReadRepresentation rr.

JLD2.read_attr_data — Function.

read_attr_data(f::JLDFile, attr::ReadAttribute)

Read data from an attribute.

read_attr_data(f::JLDFile, attr::ReadAttribute, expected_datatype::H5Datatype,
               rr::ReadRepresentation)

Read data from an attribute, assuming a specific HDF5 datatype and ReadRepresentation. If the HDF5 datatype does not match, throws an UnsupportedFeatureException. This allows better type stability while simultaneously validating the data.

JLD2.read_attribute

JLD2.read_bytestring

JLD2.read_committed_datatype

JLD2.read_compressed_array! — Function.

read_compressed_array!(v::Array, f::JLDFile, rr, data_length::Int)

Fill the array v with the compressed contents of JLDFile f at the current position, assuming a ReadRepresentation rr and that the compressed data has length data_length.

JLD2.read_data — Function.

read_data(f::JLDFile, dataspace::ReadDataspace, datatype_class::UInt8,
          datatype_offset::Int64, data_offset::Int64[, filter_id::UInt16,
          header_offset::RelOffset, attributes::Vector{ReadAttribute}])

Read data from a file. If datatype_class is typemax(UInt8), the datatype is assumed to be committed, and datatype_offset points to the offset of the committed datatype's header. Otherwise, datatype_offset points to the offset of the datatype attribute.

JLD2.read_dataspace_message

JLD2.read_datatype_message

JLD2.read_empty

JLD2.read_field_datatypes

JLD2.read_heap_object

JLD2.read_link

JLD2.read_obj_start

JLD2.read_scalar — Function.

read_scalar(f::JLDFile, rr, header_offset::RelOffset)

Read raw data representing a scalar with read representation rr from the current position of JLDFile f. header_offset is the RelOffset of the object header, used to resolve cycles.

JLD2.read_size — Function.

read_size(io::IO, flags::UInt8)

Loads a variable-length size according to flags Expects that the first two bits of flags mean: 0 The size of the Length of Link Name field is 1 byte. 1 The size of the Length of Link Name field is 2 bytes. 2 The size of the Length of Link Name field is 4 bytes. 3 The size of the Length of Link Name field is 8 bytes. Returns the size as an Int

JLD2.readmore!

JLD2.reconstruct_bitstype

JLD2.reconstruct_compound

JLD2.reconstruct_odr

JLD2.replace_expr

JLD2.samelayout

JLD2.save

JLD2.save_group — Function.

save_group(g::Group) -> RelOffset

Stores a group to a file, updating it if it has already been saved. Returns UNDEFINED_ADDRESS if the group was already stored, or the offset of the new group otherwise.

JLD2.show_group

JLD2.size_flag

JLD2.size_size

JLD2.store_vlen!

JLD2.symbol_length — Function.

symbol_length(x::Symbol)

Returns the length of the string represented by x.

JLD2.track_weakref!

JLD2.truncate_and_close

JLD2.typename

JLD2.typestring

JLD2.unsafe_isdefined

JLD2.wconvert

JLD2.write_attribute

JLD2.write_chunked_storage_message

JLD2.write_data

JLD2.write_dataset

JLD2.write_datatype_message

JLD2.write_heap_object

JLD2.write_object_header_and_dataspace_message

JLD2.write_ref

JLD2.write_ref_mutable

JLD2.write_size

JLD2.writeas