Docstrings

Yota.@diffrule

Yota.AbstractDevice

Yota.AbstractOp

Yota.Assign

Yota.CPU

Yota.Call

Yota.Constant

Yota.DEBUG_STATE

Yota.DIFF_PHS

Yota.DIFF_RULES

Yota.GPU

Yota.GRAD_CACHE

Yota.GradResult

Yota.Input

Yota.MaybeFunction

Yota.PRIMITIVES

Yota.Tape

Yota.TraceCtx

Yota.Yota

Yota._grad

Yota.add_diff_rule

Yota.back! — Function.

Backpropagate through the tape, record derivatives as new operations

Yota.check_deriv_sizes — Function.

For each input that has a derivative on this tape check if the derivative has the same size as the input.

Yota.compile

Yota.compile!

Yota.copy_with — Function.

Copy struct x replacing specified fields with new values

Yota.deriv!

Yota.deriv_broadcast!

Yota.deriv_expr — Function.

Rewrite single call expression into its derivative. Example:

deriv_expr(:(sin(x)), [Float64], 1)
# ==> :((cos(x) * ds))

Yota.device_op

Yota.deviceof

Yota.eval

Yota.exec!

Yota.field_paths

Yota.find_rules_for — Function.

Internal function for finding rule by function name

Yota.generate_body — Function.

Main part of generated function code

Yota.generate_epilogue — Function.

Generate last block of function code

Yota.generate_function_expr

Yota.generate_prologue — Function.

Generate initialization block of function code

Yota.get_arg_names

Yota.get_arg_types

Yota.getderiv

Yota.getfield_nested

Yota.grad — Function.

Find gradient of f w.r.t. its arguments. Example:

val, g = grad(sum, rand(3))

where:

val is the value of f at this point

g::GradResult is a collection of gradients

GradResult is indexed by argument index and contains gradients in a format most suitable for that argument, namely:

for arrays: arrays of the same type and size

for reals: reals

for mutable structs: dictionary of {(:field, :path) => value} pairs.

All gradients can be applied to original variables using update!() function.

Yota.guess_device

Yota.include

Yota.is_cuarray — Function.

Check if the argument is of type CuArray. Doesn't require CuArrays.jl to be loaded

Yota.isparameters

Yota.make_name

Yota.match_rule

Yota.mean_grad

Yota.minus_updater

Yota.mod

Yota.play!

Yota.record! — Function.

Record a new operation (defined by its type and arguments for constructor) to a tape and return ID of this new op.

Yota.record_expr! — Function.

Parse a complex call expression and record corresponding operations to a tape.

Optionally takes substitution table (st parameter) to replace known symbols with provided values.

Keyword params:

st::Dict - substitution table, used to replace symbols in ex with tape op ids

bcast::Bool - replace all function calls with corresponding broadcasting

Yota.recover_broadcast — Function.

Recover broadcast operation from Broadcast.broadcasted and Broadcast.materialize

Yota.replace_in_args! — Function.

Replace all call arguments according to the provided dict

Yota.resolve_functions_and_types!

Yota.resolve_old_broadcast

Yota.rewrite_mul — Function.

Rewrite :(Z = X * Y) into :(mul!(Z, X, Y)), but only if X and Y are arrays

Yota.rewrite_with_keywords

Yota.setderiv!

Yota.setfield_nested!

Yota.simplify

Yota.squash_assigned

Yota.step_back!

Yota.sum_dropdims

Yota.sum_grad

Yota.to_device

Yota.to_exgraph

Yota.to_exnode

Yota.to_unbroadcast_expr

Yota.trace — Function.

Trace function execution using provided arguments. Returns calculated value and a tape.

foo(x) = 2.0x + 1.0
val, tape = trace(foo, 4.0)

Yota.ungetindex

Yota.ungetindex!

Yota.unmake_name

Yota.update!

Yota.version

Yota.with_free_args_as_constants

Yota.without_types