Docstrings

Unrolled.@code_unrolled

Unrolled.@fixed_range — Macro.

@fixed_range 3:10 behaves like the standard range 3:10, but is stored within the type system, so that some_tuple[@fixed_range 3:10] is type-stable. Also supports @fixed_range some_tuple[3:end-5]

Unrolled.@unroll

Unrolled.@unroll_loop

Unrolled.FixedEnd — Type.

FixedRange{2, FixedEnd()}() behaves like a type-stable 2:end

Unrolled.FixedRange — Type.

FixedRange{A, B}() is like UnitRange{Int64}(A, B), but the bounds are encoded in the type.

Unrolled.Unrolled

Unrolled._unrolled_filter

Unrolled.eval

Unrolled.expansion_funs

Unrolled.function_argument_name — Function.

function_argument_name(arg_expr)

Returns the name (as a symbol) of this argument, where arg_expr is whatever can be put in a function definition's argument list (eg. len::Int=5)

Unrolled.include

Unrolled.replace_end

Unrolled.type_length — Function.

type_length(::Type) returns the length of sequences of that type (only makes sense for sequence-like types, obviously.

Unrolled.unrolled_all

Unrolled.unrolled_all_unrolled_expansion_

Unrolled.unrolled_any

Unrolled.unrolled_any_unrolled_expansion_

Unrolled.unrolled_filter

Unrolled.unrolled_foreach

Unrolled.unrolled_foreach_unrolled_expansion_

Unrolled.unrolled_in

Unrolled.unrolled_in_unrolled_expansion_ — Function.

unrolled_in(obj, tup) is like in. Beware that its return type is not always known - see #21322

source

Unrolled.unrolled_intersect

Unrolled.unrolled_map

Unrolled.unrolled_reduce

Unrolled.unrolled_setdiff

Unrolled.unrolled_union