Docstrings

SumProductNetworks.+

SumProductNetworks.-

SumProductNetworks.AbstractMixtureModel

SumProductNetworks.AbstractMvNormal

SumProductNetworks.Arcsine

SumProductNetworks.Bernoulli

SumProductNetworks.Beta

SumProductNetworks.BetaBinomial

SumProductNetworks.BetaPrime

SumProductNetworks.Binomial

SumProductNetworks.Biweight

SumProductNetworks.Categorical

SumProductNetworks.Cauchy

SumProductNetworks.Chi

SumProductNetworks.Chisq

SumProductNetworks.Continuous

SumProductNetworks.ContinuousDistribution

SumProductNetworks.ContinuousMatrixDistribution

SumProductNetworks.ContinuousMultivariateDistribution

SumProductNetworks.ContinuousUnivariateDistribution

SumProductNetworks.Cosine

SumProductNetworks.DiagNormal

SumProductNetworks.DiagNormalCanon

SumProductNetworks.Dirichlet

SumProductNetworks.DirichletMultinomial

SumProductNetworks.Discrete

SumProductNetworks.DiscreteDistribution

SumProductNetworks.DiscreteMatrixDistribution

SumProductNetworks.DiscreteMultivariateDistribution

SumProductNetworks.DiscreteUniform

SumProductNetworks.DiscreteUnivariateDistribution

SumProductNetworks.Distribution

SumProductNetworks.Distributions

SumProductNetworks.DoubleExponential

SumProductNetworks.EdgeworthMean

SumProductNetworks.EdgeworthSum

SumProductNetworks.EdgeworthZ

SumProductNetworks.EmpiricalUnivariateDistribution

SumProductNetworks.Epanechnikov

SumProductNetworks.Erlang

SumProductNetworks.Estimator

SumProductNetworks.Exponential

SumProductNetworks.FDist

SumProductNetworks.FiniteAugmentedProductNode

SumProductNetworks.FiniteProductNode

SumProductNetworks.FiniteSplitNode

SumProductNetworks.FiniteSumNode

SumProductNetworks.FisherNoncentralHypergeometric

SumProductNetworks.Frechet

SumProductNetworks.FullNormal

SumProductNetworks.FullNormalCanon

SumProductNetworks.Gamma

SumProductNetworks.GeneralizedExtremeValue

SumProductNetworks.GeneralizedPareto

SumProductNetworks.Geometric

SumProductNetworks.Gumbel

SumProductNetworks.Hypergeometric

SumProductNetworks.IndicatorNode

SumProductNetworks.InverseGamma

SumProductNetworks.InverseGaussian

SumProductNetworks.InverseWishart

SumProductNetworks.IsoNormal

SumProductNetworks.IsoNormalCanon

SumProductNetworks.KSDist

SumProductNetworks.KSOneSided

SumProductNetworks.Kolmogorov

SumProductNetworks.Laplace

SumProductNetworks.Leaf

SumProductNetworks.Levy

SumProductNetworks.LocationScale

SumProductNetworks.LogNormal

SumProductNetworks.Logistic

SumProductNetworks.MLEstimator

SumProductNetworks.MatrixDistribution

SumProductNetworks.Matrixvariate

SumProductNetworks.MixtureModel

SumProductNetworks.Multinomial

SumProductNetworks.Multivariate

SumProductNetworks.MultivariateDistribution

SumProductNetworks.MultivariateMixture

SumProductNetworks.MultivariateNode

SumProductNetworks.MultivariateNormal

SumProductNetworks.MvLogNormal

SumProductNetworks.MvNormal

SumProductNetworks.MvNormalCanon

SumProductNetworks.MvNormalKnownCov

SumProductNetworks.MvTDist

SumProductNetworks.NegativeBinomial

SumProductNetworks.Node

SumProductNetworks.NonMatrixDistribution

SumProductNetworks.NoncentralBeta

SumProductNetworks.NoncentralChisq

SumProductNetworks.NoncentralF

SumProductNetworks.NoncentralHypergeometric

SumProductNetworks.NoncentralT

SumProductNetworks.Normal

SumProductNetworks.NormalCanon

SumProductNetworks.NormalInverseGaussian

SumProductNetworks.Pareto

SumProductNetworks.Poisson

SumProductNetworks.PoissonBinomial

SumProductNetworks.ProductNode

SumProductNetworks.QQPair

SumProductNetworks.Rayleigh

SumProductNetworks.RealInterval

SumProductNetworks.SPNNode

SumProductNetworks.Sampleable

SumProductNetworks.Semicircle

SumProductNetworks.Skellam

SumProductNetworks.Statistics

SumProductNetworks.SufficientStats

SumProductNetworks.SumNode

SumProductNetworks.SumProductNetwork

SumProductNetworks.SumProductNetworks

SumProductNetworks.SymTriangularDist

SumProductNetworks.TDist

SumProductNetworks.TriangularDist

SumProductNetworks.Triweight

SumProductNetworks.Truncated

SumProductNetworks.TruncatedNormal

SumProductNetworks.Uniform

SumProductNetworks.Univariate

SumProductNetworks.UnivariateDistribution

SumProductNetworks.UnivariateGMM

SumProductNetworks.UnivariateMixture

SumProductNetworks.UnivariateNode

SumProductNetworks.ValueSupport

SumProductNetworks.VariateForm

SumProductNetworks.VonMises

SumProductNetworks.VonMisesFisher

SumProductNetworks.WalleniusNoncentralHypergeometric

SumProductNetworks.Weibull

SumProductNetworks.Wishart

SumProductNetworks.ZeroMeanDiagNormal

SumProductNetworks.ZeroMeanDiagNormalCanon

SumProductNetworks.ZeroMeanFullNormal

SumProductNetworks.ZeroMeanFullNormalCanon

SumProductNetworks.ZeroMeanIsoNormal

SumProductNetworks.ZeroMeanIsoNormalCanon

SumProductNetworks.add! — Function.

Add a node to a finite sum node with given weight in place. add!(node::FiniteSumNode, child::SPNNode, weight<:Real)

SumProductNetworks.addobs!

SumProductNetworks.addscope!

SumProductNetworks.binaryentropy

SumProductNetworks.bmitest — Function.

Test if discrete sets are independent.

test(X, Y) -> p

Computes the Independence Test by Margaritis and Thrun for X and Y.
The test returns the probability of X and Y being independent.

SumProductNetworks.canonform

SumProductNetworks.ccdf

SumProductNetworks.cdf

SumProductNetworks.cf

SumProductNetworks.cgf

SumProductNetworks.children — Function.

children(node) -> children::SPNNode[]

Returns the children of an internal node.

Parameters:

node::Node: Internal SPN node to be evaluated.

SumProductNetworks.circvar

SumProductNetworks.classes

SumProductNetworks.complexity — Function.

complexity(S)

Compute the complexity (number of free parameters) of the SPN rooted at S.

SumProductNetworks.component

SumProductNetworks.components

SumProductNetworks.componentwise_logpdf

SumProductNetworks.componentwise_pdf

SumProductNetworks.concentration

SumProductNetworks.cor

SumProductNetworks.cov

SumProductNetworks.cquantile

SumProductNetworks.cumulant

SumProductNetworks.depth — Function.

depth(S)

Compute the depth of the SPN rooted at S.

SumProductNetworks.dim

SumProductNetworks.dof

SumProductNetworks.entropy

SumProductNetworks.estimate

SumProductNetworks.eval

SumProductNetworks.expected_logdet

SumProductNetworks.export_network

SumProductNetworks.failprob

SumProductNetworks.fit

SumProductNetworks.fit_map

SumProductNetworks.fit_map!

SumProductNetworks.fit_mle

SumProductNetworks.fit_mle!

SumProductNetworks.freecumulant

SumProductNetworks.generate_spn — Function.

generate_spn(X::Matrix, algo::Symbol; ...)

Generate an SPN structure using a structure learning algorithm.

Arguments:

X: Data matrix.

algo: Algorithm, one out of [:learnspn, :random].

SumProductNetworks.getFontSize

SumProductNetworks.getLabel

SumProductNetworks.getOrderedNodes — Function.

Get all nodes in topological order using Tarjan's algoritm.

SumProductNetworks.getParameters

SumProductNetworks.getShape

SumProductNetworks.gradlogpdf

SumProductNetworks.hasfinitesupport

SumProductNetworks.hasobs

SumProductNetworks.hasscope

SumProductNetworks.hassubscope

SumProductNetworks.hasweights

SumProductNetworks.header

SumProductNetworks.include

SumProductNetworks.insupport

SumProductNetworks.invcov

SumProductNetworks.invlogccdf

SumProductNetworks.invlogcdf

SumProductNetworks.invscale

SumProductNetworks.isDegenerated

SumProductNetworks.isbounded

SumProductNetworks.isleptokurtic

SumProductNetworks.islowerbounded

SumProductNetworks.ismesokurtic

SumProductNetworks.isnormalized

SumProductNetworks.isplatykurtic

SumProductNetworks.isprobvec

SumProductNetworks.isupperbounded

SumProductNetworks.kde

SumProductNetworks.kurtosis

SumProductNetworks.learnspn

SumProductNetworks.length

SumProductNetworks.location

SumProductNetworks.location!

SumProductNetworks.logccdf

SumProductNetworks.logcdf

SumProductNetworks.logdetcov

SumProductNetworks.loglikelihood

SumProductNetworks.logpdf

SumProductNetworks.logpdf!

SumProductNetworks.logweights

SumProductNetworks.mean

SumProductNetworks.mean!

SumProductNetworks.meandir

SumProductNetworks.meanform

SumProductNetworks.meanlogx

SumProductNetworks.median

SumProductNetworks.median!

SumProductNetworks.mgf

SumProductNetworks.middle

SumProductNetworks.mode

SumProductNetworks.modes

SumProductNetworks.moment

SumProductNetworks.ncategories

SumProductNetworks.ncomponents

SumProductNetworks.ndepth

SumProductNetworks.nobs

SumProductNetworks.normalize! — Function.

normalize!(S)

Localy normalize the weights of a SPN using Algorithm 1 from Peharz et al.

SumProductNetworks.nsamples

SumProductNetworks.nscope

SumProductNetworks.ntrials

SumProductNetworks.obs

SumProductNetworks.params

SumProductNetworks.params!

SumProductNetworks.parents — Function.

parents(node) -> parents::SPNNode[]

Returns the parents of an SPN node.

Parameters:

node::SPNNode: SPN node to be evaluated.

SumProductNetworks.partype

SumProductNetworks.pdf

SumProductNetworks.probs

SumProductNetworks.probval

SumProductNetworks.qqbuild

SumProductNetworks.quantile

SumProductNetworks.quantile!

SumProductNetworks.randomspn

SumProductNetworks.rate

SumProductNetworks.remove! — Function.

Remove a node from the children list of a sum node in place. remove!(node::FiniteSumNode, index::Int)

SumProductNetworks.removescope!

SumProductNetworks.sample

SumProductNetworks.sample!

SumProductNetworks.sampler

SumProductNetworks.scale

SumProductNetworks.scale!

SumProductNetworks.scope

SumProductNetworks.setobs!

SumProductNetworks.setscope!

SumProductNetworks.shape

SumProductNetworks.skewness

SumProductNetworks.span

SumProductNetworks.sqmahal

SumProductNetworks.sqmahal!

SumProductNetworks.std

SumProductNetworks.stdlogx

SumProductNetworks.stdm

SumProductNetworks.sub2ind2 — Function.

sub2ind2(size1, ind1, ind2) -> linear index

SumProductNetworks.sub2ind3 — Function.

sub2ind3(size1, size2, ind1, ind2, ind3) -> linear index

SumProductNetworks.succprob

SumProductNetworks.suffstats

SumProductNetworks.support

SumProductNetworks.test_distr

SumProductNetworks.test_samples

SumProductNetworks.updatescope! — Function.

updatescope!(spn)

Update the scope of all nodes in the SPN.

SumProductNetworks.var

SumProductNetworks.varlogx

SumProductNetworks.varm

SumProductNetworks.visitNode!

SumProductNetworks.weights

SumProductNetworks.wsample

SumProductNetworks.wsample!