Docstrings

Libtask.Libtask

Libtask.TArray — Type.

TArray{T}(dims, ...)

Implementation of data structures that automatically perform copy-on-write after task copying.

If currenttask is an existing key in s, then return `s[currenttask]. Otherwise, returns[currenttask] = s[lasttask]`.

Usage:

TArray(dim)

Example:

ta = TArray(4)              # init
for i in 1:4 ta[i] = i end  # assign
Array(ta)                   # convert to 4-element Array{Int64,1}: [1, 2, 3, 4]

Libtask.check_deps

Libtask.consume

Libtask.eval

Libtask.get

Libtask.include

Libtask.libtask

Libtask.n_copies

Libtask.produce

Libtask.tzeros — Function.

 tzeros(dims, ...)

Construct a distributed array of zeros. Trailing arguments are the same as those accepted by TArray.

tzeros(dim)

Example:

tz = tzeros(4)              # construct
Array(tz)                   # convert to 4-element Array{Int64,1}: [0, 0, 0, 0]