Docstrings

SpatialIndexing.AbstractPool — Type.

Base class for implementating the pool of T objects. The pool allows to reduce the stress on GC by collecting the unneeded objects (release!(pool, obj)) and reusing them later (acquire!(pool)).

SpatialIndexing.Branch — Type.

R-Tree node for levels above 1 (non-Leaf).

SpatialIndexing.BranchPool

SpatialIndexing.HasID — Type.

Type trait for checking id() method support. If type V has this trait (idtype(V) returns HasID{K}), then id(v::V) should return a unique identifier for v of type K. If V doesn't have this trait, idtype(V) returns HasNoID.

If available, SpatialIndex{T,N,V} uses unique identifiers of V alongside spatial indexing.

SpatialIndexing.HasMBR — Type.

Type trait for checking mbr() method support. If type V has this trait (mbrtype(V) returns HasMBR{Rect{T,N}}), then mbr(v::V) should return a minimal bounding rectangle (MBR) Rect{T,N} that contains v. If V doesn't have this trait, mbrtype(V) returns HasNoMBR.

SpatialIndex{T,N,V} requires that V provides mbr() method that returns Rect{T,N}.

SpatialIndexing.HasNoID

SpatialIndexing.HasNoMBR

SpatialIndexing.Leaf — Type.

R-Tree leaf (level 1 node). Its children are data elements of type V.

SpatialIndexing.LeafPool

SpatialIndexing.Node — Type.

Base class for RTree node (Branch or Leaf).

SpatialIndexing.NodePool — Type.

Pool of the T1 R-tree nodes (Leaf or Branch) with T2 children (Branch, Leaf or Elem). It allows reusing the deleted nodes and reduce the stress on GC.

SpatialIndexing.Point — Type.

N-dimensional point.

SpatialIndexing.Pool — Type.

The default AbstarctPool implementation.

SpatialIndexing.QueryContainedIn

SpatialIndexing.QueryIntersectsWith

SpatialIndexing.QueryKind — Type.

Specifies the kind of spatial data query.

source

SpatialIndexing.QueryMatch — Type.

Specifies the result of spatial data query.

source

SpatialIndexing.QueryMatchComplete

SpatialIndexing.QueryMatchPartial

SpatialIndexing.QueryNoMatch

SpatialIndexing.RTree — Type.

R-Tree: N-dimensional spatial data index [guttman84].

R-tree groups data elements (V) into leaves (Leaf) and leaves into branches (Branch). It uses various heuristics to ensure that the minimal bounding rectangles (MBRs) of the nodes (Rect{T,N} rectangles that encompass the data elements attached to these nodes) stay compact and that the MBRs of the nodes that are on the same level of R-tree hierarchy have minimal overlap with each other. This property makes R-trees efficient for spatial queries.

To facilitate spatial indexing, the V data elements need to support HasMBR trait (i.e. define mbrtype(V) and mbr(v::V) methods) and, optionally, HasID trait (via idtype(V) and id(v::V) methods). mbr(v::V) should return minimal bounding rectangle (MBR) of type Rect{T,N} that contains v. SpatialElem{T,N,D} type provides generic implementation of spatial data element that explicitly stores id, mbr and data object of type D and implements HasMBR and HasID traits.

Parameters

The behaviour of RTree is defined by the parameters supplied at its creation:

T: the numeric type for the spatial coordinate

N: the number of spatial dimensions

variant: one of RTreeLinear, RTreeQuadratic, or RTreeStar (default)

tight_mbrs: recalculate node MBR when the child is removed (default is true)

branch_capacity: capacity of branch nodes (default is 100)

leaf_capacity: capacity of leaf nodes (default is 100)

leafpool_capacity: How many detached 1st level nodes (leaves) should be kept for reuse (default is 100)

twigpool_capacity: How many detached 2nd level nodes should be kept for reuse (default is 100)

branchpool_capacity: How many other (level > 2) detached branch nodes should be kept for reuse (default is 100)

nearmin_overlap: How many candidates to consider when identifying the node with minimal overlap (default is 32)

fill_factor: How much should the node be filled (fraction of its capacity) after splitting (default is 0.7)

split_factor: How much can the sizes of the two nodes differ after splitting (default is 0.4)

reinsert_factor: How much should the node be underfilled (fraction of its capacity) to consider removing it and redistributing its children to other nodes (default is 0.3)

Performance

The nodes in R-tree have limited capacity (maximual number of children) specified at RTree creation (leaf_capacity and branch_capacity). Larger capacities results in shorter trees, but they time required to locate the specific spatial region grows linearly with the capacity.

References

[guttman84] “R-Trees: A Dynamic Index Structure for Spatial Searching” A. Guttman, Proc. 1984 ACM-SIGMOD Conference on Management of Data (1985), 47-57. [beckmann90] "The R*-tree: an efficient and robust access method for points and rectangles" N. Beckmann, H.P. Kriegel, R. Schneider, B. Seeger, Proc. 1990 ACM SIGMOD international conference on Management of data (1990), p.322

SpatialIndexing.RTreeIteratorState

SpatialIndexing.RTreeLinear

SpatialIndexing.RTreeQuadratic

SpatialIndexing.RTreeQueryIteratorState

SpatialIndexing.RTreeRegionQueryIterator

SpatialIndexing.RTreeStar

SpatialIndexing.RTreeVariant — Type.

R-Tree variants.

source

SpatialIndexing.Rect — Type.

Rectangular region constrained by low[i]...high[i] in each of N dimensions.

SpatialIndexing.Region — Type.

Base abstract class for implementing regions in N-dimensional space with dimensions of type T.

SpatialIndexing.SimpleSpatialIndex — Type.

Vector-based SpatialIndex. While insertion is O(1), the search is O(N).

Generally should not be used, except for performance comparisons or when the number of stored elements is expected to be very small (<100).

SpatialIndexing.SpatialElem — Type.

Simple N-dimensional spatial data element that stores values of type V and supports HasMBR{Rect{T,N}} and HasID{K} (if K is not Nothing) traits.

SpatialIndexing.SpatialIndex — Type.

Base abstract class for spatial indexing of elements of type V in N-dimensional space with dimensions of type T.

SpatialIndexing.SpatialIndexException — Type.

SpatialIndex-related exception raised within Julia

SpatialIndexing.SpatialIndexing

SpatialIndexing.SpatialQueryIterator — Type.

Base abstract class for implementing spatial queries in N-dimensional space.

SpatialIndexing.SubtreeContext

SpatialIndexing.TwigPool

SpatialIndexing.__spatial_keyerror

SpatialIndexing._attach!

SpatialIndexing._condense!

SpatialIndexing._detach!

SpatialIndexing._insert!

SpatialIndexing._insert!_fullnode

SpatialIndexing._insert!_fullnode_rstar

SpatialIndexing._isempty

SpatialIndexing._iterate

SpatialIndexing._mbr_high

SpatialIndexing._mbr_low

SpatialIndexing._nextchild

SpatialIndexing._reinsert!

SpatialIndexing._release_descendants!

SpatialIndexing._replace!

SpatialIndexing._split!

SpatialIndexing._split!_rstar

SpatialIndexing._split!_rtree

SpatialIndexing._splitseeds

SpatialIndexing._splitseeds_linear

SpatialIndexing._splitseeds_quadratic

SpatialIndexing._subtract!

SpatialIndexing._updatembr!

SpatialIndexing.acquire

SpatialIndexing.acquire!

SpatialIndexing.area — Function.

area(a::Region{T,N}) where {T,N}

N-dimensional "area" (volume etc) of a.

SpatialIndexing.branchtype

SpatialIndexing.capacity

SpatialIndexing.center

SpatialIndexing.check

SpatialIndexing.check_eltype_rtree

SpatialIndexing.check_hasid

SpatialIndexing.check_hasmbr

SpatialIndexing.children

SpatialIndexing.choose_subtree

SpatialIndexing.combine — Function.

combine(a::Region{T,N}, b::Region{T,N}) where {T,N}

MBR that contains both a and b regions.

SpatialIndexing.combined_area

SpatialIndexing.contained_in — Function.

contained_in(index::SpatialIndex, region::Region)

Get iterator for index elements contained in region.

SpatialIndexing.contains — Function.

contains(a::Region{T,N}, b::Region{T,N}) where {T,N}

Check whether a contains b.

SpatialIndexing.delete_subtree!

SpatialIndexing.dimtype

SpatialIndexing.empty — Function.

empty(::Type{T}) where T<:Region

Generate empty (uninitialized) region of type T.

SpatialIndexing.enlargement

SpatialIndexing.eval

SpatialIndexing.find_least_enlargement

SpatialIndexing.find_least_overlap

SpatialIndexing.hasparent

SpatialIndexing.height

SpatialIndexing.id

SpatialIndexing.idtrait

SpatialIndexing.idtype

SpatialIndexing.include

SpatialIndexing.intersect

SpatialIndexing.intersects — Function.

intersects(a::Region{T,N}, b::Region{T,N}) where {T,N}

Check whether a intersects with b.

SpatialIndexing.intersects_with — Function.

intersects_with(index::SpatialIndex, region::Region)

Get iterator for index elements intersecting with region.

SpatialIndexing.isequal_rtree

SpatialIndexing.isoverflow

SpatialIndexing.isvalid — Function.

isvalid(a::Region)

Check that the parameters of a are valid and it defines a proper region.

SpatialIndexing.leaftype

SpatialIndexing.level

SpatialIndexing.load! — Function.

load!(tree::RTree{T,N,V}, data::Any;
      convertel = identity, method = :OMT,
      leaf_fill = capacity(Leaf, tree),
      branch_fill::Tuple{Integer, Integer} = omt_branch_fill(tree)) where {T,N,V}

Bulk-load data into tree.

tree: an empty R-tree for storing elements of type V

data: iterable container with the elements to put into tree

convertel: function to convert elements of data to type V

method: bulk-loading method

leaf_fill: the average number of elements to store in R-tree leaves (1-level nodes)

branch_fill: the tuple of the number of slices and the number of subtrees per slice in the R-tree nodes (level ≥ 1).

The supported bulk-loading methods are:

:OMT: Overlap Minimizing Top-down method by Taewon Lee and Sukho Lee

SpatialIndexing.load_omt!

SpatialIndexing.mbr

SpatialIndexing.mbrtrait

SpatialIndexing.mbrtype

SpatialIndexing.nelements

SpatialIndexing.newelem

SpatialIndexing.node_capacity

SpatialIndexing.nodetype

SpatialIndexing.omt_branch_fill

SpatialIndexing.omt_subtree

SpatialIndexing.overlap_area

SpatialIndexing.parent

SpatialIndexing.perimeter — Function.

perimeter(a::Region)

The sum of the a sides.

SpatialIndexing.pos_in_parent

SpatialIndexing.querykind

SpatialIndexing.regiontype

SpatialIndexing.release

SpatialIndexing.release!

SpatialIndexing.setoverflow!

SpatialIndexing.should_visit

SpatialIndexing.sqrdistance

SpatialIndexing.subtract! — Function.

subtract!(tree::RTree, reg::Region)

Subtracts the region from the tree, i.e. removes all elements within region.

SpatialIndexing.syncmbr!

SpatialIndexing.touches — Function.

touches(a::Rect{T,N}, b::Rect{T,N}) where {T,N}

Check whether a and b touch (i.e. any low side touches low or high touches high).

SpatialIndexing.variant