Docstrings

BitBasis.@bit_str — Macro.

@bit_str -> BitStr

Construct a bit string. such as bit"0000". The bit strings also supports string bcat. Just use it like normal strings.

Example

julia> bit"10001"
10001 (17)

julia> bit"100_111_101"
00001110101 (117)

julia> bcat(bit"1001", bit"11", bit"1110")
1001111110 (638)

julia> v = rand(16);

julia> v[bit"1001"]
0.38965443157314406

julia> onehot(bit"1001")
16-element Array{Float64,1}:
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 1.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0

BitBasis.BitBasis

BitBasis.BitStr — Type.

BitStr{T}

String literal for bits.

BitStr(value[, len=ndigits(value)])

Returns a BitStr, by default the length is set to the minimum length required to represent value as bits.

BitStr(str::String)

Parse the input string to a BitStr. See @bit_str for more details.

Example

BitStr supports some basic arithmetic operations. It acts like an integer, but supports some frequently used methods for binary basis.

julia> bit"101" * 2
1010 (10)

julia> bcat(bit"101" for i in 1:10)
101101101101101101101101101101 (766958445)

julia> repeat(bit"101", 2)
101101 (45)

julia> bit"1101"[2]
0

BitBasis.IntIterator

BitBasis.ReorderedBasis — Type.

ReorderedBasis{N, T}

Lazy reorderd basis.

BitBasis._packbits

BitBasis._reorder

BitBasis.allmasked — Function.

allmasked(index::Integer, mask::Integer) -> Bool

Return true if all masked position of index is 1.

BitBasis.anymasked — Function.

anymasked(index::Integer, mask::Integer) -> Bool

Return true if any masked position of index is 1.

BitBasis.baddrs — Function.

baddrs(b::Integer) -> Vector

get the locations of nonzeros bits, i.e. the inverse operation of bmask.

BitBasis.basis — Function.

basis([IntType], num_bits::Int) -> UnitRange{IntType}
basis([IntType], state::AbstractArray) -> UnitRange{IntType}

Returns the UnitRange for basis in Hilbert Space of num_bits qubits. If an array is supplied, it will return a basis having the same size with the first diemension of array.

BitBasis.bcat

BitBasis.bdistance — Function.

bdistance(i::Integer, j::Integer) -> Int

Return number of different bits.

BitBasis.bfloat — Function.

bfloat(b::Integer; nbit::Int=bit_length(b)) -> Float64

float view, with big end qubit 1.

BitBasis.bfloat_r — Function.

bfloat_r(b::Integer; nbit::Int) -> Float64

float view, with bits read in inverse order.

BitBasis.bint — Function.

bint(b; nbit=nothing) -> Int

integer view, with little end qubit 1.

BitBasis.bint_r — Function.

bint_r(b; nbit::Int) -> Integer

integer read in inverse order.

BitBasis.bit_length — Function.

bit_length(x::Integer) -> Int

Return the number of bits required to represent input integer x.

BitBasis.bitarray — Function.

bitarray(v::Vector, [num_bits::Int]) -> BitArray
bitarray(v::Int, num_bits::Int) -> BitArray
bitarray(num_bits::Int) -> Function

Construct BitArray from an integer vector, if num_bits not supplied, it is 64. If an integer is supplied, it returns a function mapping a Vector/Int to bitarray.

BitBasis.bmask — Function.

bmask(::Type{T}) where T <: Integer -> zero(T)
bmask([T::Type], positions::Int...) -> T
bmask([T::Type], range::UnitRange{Int}) -> T

Return an integer mask of type T where 1 is the position masked according to positions or range. Directly use T will return an empty mask 0.

BitBasis.breflect — Function.

breflect(num_bits::Int, b::Integer[, masks::Vector{Integer}]) -> Integer

Return left-right reflected integer.

BitBasis.bsizeof — Function.

bsizeof(::Type)

Returns the size of given type in number of binary digits.

BitBasis.controller — Function.

controller(cbits, cvals) -> Function

Return a function that checks whether a basis at cbits takes specific value cvals.

BitBasis.eat_underscore

BitBasis.eval

BitBasis.flip — Function.

flip(index::Integer, mask::Integer) -> Integer

Return an Integer with bits at masked position flipped.

BitBasis.getbit

BitBasis.hypercubic — Function.

hypercubic(A::Array) -> Array

get the hypercubic representation for an array.

BitBasis.include

BitBasis.int

BitBasis.invorder — Function.

invorder(X::AbstractVecOrMat)

Inverse the order of given vector/matrix X.

BitBasis.ismasked_equal — Function.

ismasked_equal(index::Integer, mask::Integer, onemask::Integer) -> Bool

Return true if bits at positions masked by mask equal to 1 are equal to onemask.

Example

julia> n = 0b11001; mask = 0b10100; onemask = 0b10000;

julia> ismasked_equal(n, mask, onemask)
true

BitBasis.log2dim1 — Function.

log2dim1(X)

Returns the log2 of the first dimension's size.

BitBasis.log2i — Function.

log2i(x::Integer) -> Integer

Return log2(x), this integer version of log2 is fast but only valid for number equal to 2^n.

BitBasis.neg — Function.

neg(index::Integer, num_bits::Int) -> Integer

Return an integer with all bits flipped (with total number of bit num_bits).

BitBasis.next_reordered_basis — Function.

next_reordered_basis(basis, takers, differ)

Returns the next reordered basis accroding to current basis.

BitBasis.onehot — Function.

onehot([T=Float64], nbits, x::Integer)

Returns an onehot vector of type Vector{T}, where index x + 1 is one.

onehot([T=Float64], bit_str)

Returns an onehot vector of type Vector{T}, where the bit_str-th element is one.

BitBasis.packbits — Function.

packbits(arr::AbstractArray) -> AbstractArray

pack bits to integers, usually take a BitArray as input.

BitBasis.parse_bit

BitBasis.reorder — Function.

reorder(X::AbstractArray, orders)

Reorder X according to orders.

Tip

Although orders can be any iterable, Tuple is preferred inorder to gain as much performance as possible. But the conversion won't take much anyway.

BitBasis.setbit — Function.

setbit(index::Integer, mask::Integer) -> Integer

set the bit at masked position to 1.

BitBasis.sum_length

BitBasis.swapbits — Function.

swapbits(n::Integer, mask_ij::Integer) -> Integer

Return an integer with bits at i and j in given mask flipped.

Warning

mask_ij should only contain two 1, swapbits will not check it, use at your own risk.

swapbits(n::Integer, i::Int, j::Int) -> Integer

Return an integer with bits at i and j flipped.

BitBasis.takebit

BitBasis.testall

BitBasis.testany

BitBasis.testval

BitBasis.unsafe_reorder — Function.

unsafe_reorder(X::AbstractArray, orders)

Reorder X according to orders.

Warning

unsafe_reorder won't check whether the length of orders and the size of first dimension of X match, use at your own risk.

BitBasis.unsafe_sub — Function.

unsafe_sub(a::UnitRange, b::NTuple{N}) -> NTuple{N}

Returns result in type Tuple of a .- b. This will not check the length of a and b, use at your own risk.

unsafe_sub(a::UnitRange{T}, b::Vector{T}) where T

Returns a .- b, fallback version when b is a Vector.

BitBasis.unsafe_swapbits