Docstrings

JWTs.JWK — Constant.

JWK represents a JWK Key (either for signing or verification).

JWK can be either a JWKRSA or JWKSymmetric. A RSA key can represent either the public or private key.

JWTs.JWKRSA

JWTs.JWKSet — Type.

JWKSet holds a set of keys, fetched from a OpenId key URL, each key identified by a key id.

The key URL can either be of http(s):// or file:// type.

JWTs.JWKSymmetric

JWTs.JWT — Type.

JWT represents a JWT payload at the minimum.

When signed, it holds the header and signature too. The parts are stored in encoded form.

JWTs.JWTs

JWTs.claims

JWTs.decodepart

JWTs.eval

JWTs.include

JWTs.issigned

JWTs.isvalid

JWTs.isverified

JWTs.kid

JWTs.padb64

JWTs.pubkey

JWTs.refresh!

JWTs.show

JWTs.sign!

JWTs.urldec

JWTs.urlenc

JWTs.validate!