Docstrings

Inflate.AbstractInflateData

Inflate.AbstractInflateStream

Inflate.Inflate — Module.

Inflate

Pure Julia implementation of decompression of the Deflate format and the Zlib and Gzip wrapper formats.

In-memory decompression is done by the following functions:

function	decompresses

inflate(data::Vector{UInt8})	Deflate data

inflate_zlib(data::Vector{UInt8})	Zlib data

inflate_gzip(data::Vector{UInt8})	Gzip data

Streaming decompression is done using the following types:

stream	decompresses

InflateStream(stream::IO)	Deflate stream

InflateZlibStream(stream::IO)	Zlib stream

InflateGzipStream(stream::IO)	Gzip stream

Inflate.InflateData

Inflate.InflateGzipStream — Type.

InflateGzipStream(stream::IO)

Streaming decompression of Gzip compressed stream. For an in-memory counterpart, see inflate_gzip.

gzip_headers = Dict{String, Any}()
InflateGzipStream(stream::IO; headers = gzip_headers)

Also retrieve gzip headers in the provided Dict. The headers are available directly after the object is constructed.

InflateGzipStream(stream::IO; ignore_checksum = true)

Skip computing CRC of the content, as well as the header, for consistency checking.

Reference: RFC 1952

Inflate.InflateStream — Type.

InflateStream(stream::IO)

Streaming decompression of unwrapped deflate compressed stream. For an in-memory counterpart, see inflate.

Reference: RFC 1951

Inflate.InflateZlibStream — Type.

InflateZlibStream(stream::IO)

Streaming decompression of Zlib compressed stream. For an in-memory counterpart, see inflate_zlib.

InflateZlibStream(stream::IO; ignore_checksum = true)

Skip computing Adler checksum for consistency checking.

Reference: RFC 1950

Inflate.StreamingInflateData

Inflate._inflate

Inflate.base_length

Inflate.buffer_size

Inflate.compute_adler_checksum

Inflate.crc

Inflate.crc_table

Inflate.eval

Inflate.extra_length_bits

Inflate.finish_adler

Inflate.finish_crc

Inflate.fixed_distance_table

Inflate.fixed_literal_or_length_table

Inflate.get_aligned_byte

Inflate.get_input_byte

Inflate.get_input_bytes

Inflate.get_literal_or_length

Inflate.get_value_from_code

Inflate.getbit

Inflate.getbits

Inflate.getbyte

Inflate.getdist

Inflate.getlength

Inflate.include

Inflate.inflate — Function.

inflate(source::Vector{UInt8})

Decompress in-memory source, in unwrapped deflate format. The output will also be a Vector{UInt8}. For a streaming counterpart, see InflateStream.

Reference: RFC 1951

Inflate.inflate_gzip — Function.

inflate_gzip(source::Vector{UInt8})

Decompress in-memory source, in Gzip compressed format. The output will also be a Vector{UInt8}. For a streaming counterpart, see InflateGzipStream.

gzip_headers = Dict{String, Any}()
inflate_gzip(source::Vector{UInt8}; headers = gzip_headers)

Also retrieve gzip headers in the provided Dict.

inflate_gzip(source::Vector{UInt8}; ignore_checksum = true)

Skip computing CRC of the content, as well as the header, for consistency checking.

Reference: RFC 1952

inflate_gzip(filename::AbstractString)

Convenience wrapper for reading a gzip compressed text file. The result is returned as a string.

Inflate.inflate_zlib — Function.

inflate_zlib(source::Vector{UInt8})

Decompress in-memory source, in Zlib compressed format. The output will also be a Vector{UInt8}. For a streaming counterpart, see InflateZlibStream.

inflate_zlib(source::Vector{UInt8}; ignore_checksum = true)

Skip computing Adler checksum for consistency checking.

Reference: RFC 1950

Inflate.init_adler

Inflate.init_crc

Inflate.make_crc_table

Inflate.order

Inflate.read_code_tables

Inflate.read_gzip_header

Inflate.read_output_byte

Inflate.read_output_bytes!

Inflate.read_trailer

Inflate.read_zero_terminated_data

Inflate.read_zlib_header

Inflate.skip_bits_to_byte_boundary

Inflate.transform_code_lengths_to_code

Inflate.update_adler

Inflate.update_crc

Inflate.write_to_buffer