Docstrings

TikzQTrees.@qtree

TikzQTrees.AbstractTree — Type.

AbstractTree{T}

abstract type for trees with values of type T

Methods to implement

value(tree): Returns the value of the root of the tree

children(tree): Returns an iterator over the children of the root of the tree

TikzQTrees.SimpleTree

TikzQTrees.TikzQTree

TikzQTrees.TikzQTrees

TikzQTrees._tree

TikzQTrees.children — Function.

children(tree)

Returns an iterator over the children of the root of the tree

TikzQTrees.eval

TikzQTrees.include

TikzQTrees.isleaf

TikzQTrees.latex_string

TikzQTrees.leafs

TikzQTrees.value — Function.

value(tree)

Returns the value of the root of the tree