Docstrings

IncompleteLU.ILUFactorization

IncompleteLU.IncompleteLU

IncompleteLU.InsertableSparseVector — Type.

InsertableSparseVector accumulates the sparse vector result from SpMV. Initialization requires O(N) work, therefore the data structure is reused. Insertion requires O(nnz) at worst, as insertion sort is used.

IncompleteLU.LinkedLists — Type.

The factor L is stored column-wise, but we need all nonzeros in row row. We already keep track of the first nonzero in each column (at most n indices). Take l = LinkedLists(n). Let l.head[row] be the column of some nonzero in row row. Then we can store the column of the next nonzero of row row in l.next[l.head[row]], etc. That "spot" is empty and there will never be a conflict because as long as we only store the first nonzero per column: the column is then a unique identifier.

IncompleteLU.RowReader

IncompleteLU.SortedSet — Type.

SortedSet keeps track of a sorted set of integers ≤ N using insertion sort with a linked list structure in a pre-allocated vector. Requires O(N + 1) memory. Insertion goes via a linear scan in O(n) where n is the number of stored elements, but can be accelerated by passing along a known value in the set (which is useful when pushing in an already sorted list). The insertion itself requires O(1) operations due to the linked list structure. Provides iterators:

ints = SortedSet(10)
push!(ints, 5)
push!(ints, 3)
for value in ints
    println(value)
end

IncompleteLU.SparseVectorAccumulator

IncompleteLU.add! — Function.

Sets v[idx] += a when idx is occupied, or sets v[idx] = a. Complexity is O(1).

Sets v[idx] += a when idx is occupied, or sets v[idx] = a. Complexity is O(nnz). The prev_idx can be used to start the linear search at prev_idx, useful when multiple already sorted values are added.

Add without providing a previous index.

IncompleteLU.append_col! — Function.

Basically A[:, j] = scale * drop(y), where drop removes values less than drop. Note: sorts the nzind's of y, so that the column can be appended to a SparseMatrixCSC.

Resets the SparseVectorAccumulator.

Note: does not update A.colptr for columns > j + 1, as that is done during the steps.

Basically A[:, j] = scale * drop(y), where drop removes values less than drop.

Resets the InsertableSparseVector.

Note: does not update A.colptr for columns > j + 1, as that is done during the steps.

IncompleteLU.enqueue_next_nonzero!

IncompleteLU.eval

IncompleteLU.first_in_row

IncompleteLU.forward_substitution_without_diag! — Function.

Applies in-place forward substitution with the L factor, under the assumptions:

L is stored column-wise (unlike U)

L has no upper triangular elements

L has no diagonal elements

IncompleteLU.has_next_nonzero

IncompleteLU.ilu

IncompleteLU.include

IncompleteLU.indices

IncompleteLU.is_column

IncompleteLU.isoccupied — Function.

Check whether idx is nonzero.

IncompleteLU.next_column

IncompleteLU.next_row!

IncompleteLU.nzidx

IncompleteLU.nzrow

IncompleteLU.nzval

IncompleteLU.transposed_backward_substitution! — Function.

Applies in-place backward substitution with the U factor, under the assumptions:

U is stored transposed / row-wise

U has no lower-triangular elements stored

U has (nonzero) diagonal elements stored.