Docstrings

BilevelOptimization.BilevelFlowProblems — Module.

Data structure and resolution methods of some forms of bilevel flow problems

BilevelOptimization.BilevelLP — Type.

A bilevel linear optimization problem of the form:

min cx^T * x + cy^T * y
s.t. G x + H y <= q
     x_j ∈ [xl_j,xu_j]
     x_j ∈ ℤ ∀ j ∈ Jx
     y ∈ arg min {
        d^T * y + x^T * F * y
        s.t. A x + B y <= b
             y_j ∈ [yl_j,yu_j]
        }

Note that integer variables are allowed at the upper level.

BilevelOptimization.BilevelOptimization

BilevelOptimization.MT — Type.

Convenience alias for a matrix of a sub-type of Real

BilevelOptimization.VT — Type.

Convenience alias for a vector of a sub-type of Real

BilevelOptimization.VariableType — Type.

VariableType enum distinguishing upper- and lower-level variables for setting upper and lower bounds

source

BilevelOptimization.build_blp_model — Function.

Build the JuMP model from the data of bfp and assign it the passed solver

Build a JuMP model (constraints and objective) based on the data from bp and with a solver

Add the lower-level constraints and optimality conditions to an existing JuMP model. This assumes the upper-level feasibility constraints and objective have already been set

Build the bilevel JuMP model from the data without grouping everything into a BilevelLP

BilevelOptimization.eval

BilevelOptimization.include

BilevelOptimization.lower

BilevelOptimization.upper