Docstrings

Immerse.@colorant_str

Immerse.Col

Immerse.ContainersWithChildren

Immerse.Coord

Immerse.Figure — Type.

Figure(3) gets the underlying Figure object associated with figure #3. This can be useful if you need to layer on extra drawing on top of what Gadfly produces.

Immerse.FormIterator

Immerse.FormNativeIterator

Immerse.Gadfly

Immerse.GadflyDisplay

Immerse.Geom

Immerse.Guide

Immerse.HOME

Immerse.Immerse

Immerse.Iterables

Immerse.Layer

Immerse.PDF

Immerse.PGF

Immerse.PNG

Immerse.PS

Immerse.PanZoomCallbacks

Immerse.Plot

Immerse.Row

Immerse.SVG

Immerse.SVGJS

Immerse.Scale

Immerse.Shape

Immerse.Stat

Immerse.Theme

Immerse.__init__

Immerse._aes

Immerse._coord

Immerse._display

Immerse._find_parent

Immerse._getvalue

Immerse._hit_data

Immerse._plot

Immerse._setproperty!

Immerse.absolute_to_data

Immerse.addfig

Immerse.apply_tweaks!

Immerse.bareobj

Immerse.block

Immerse.circle

Immerse.circle_center

Immerse.clear_guidata

Immerse.clear_guistate!

Immerse.clear_hit

Immerse.closeall — Function.

closeall() closes all existing figure windows.

Immerse.closefig — Function.

closefig(n) closes the nth figure window.

closefig() closes the current figure window.

Immerse.cm

Immerse.color

Immerse.cons

Immerse.createPlotGuiComponents

Immerse.cross

Immerse.curfig

Immerse.dev2data

Immerse.diamond

Immerse.disconnect

Immerse.done

Immerse.dragging

Immerse.draw

Immerse.dropfig

Immerse.dtriangle

Immerse.eval

Immerse.export_selection

Immerse.figure — Function.

figure(;name="Figure $n", width=400, height=400) creates a new figure window for displaying plots.

figure(n) raises the nth figure window and makes it the current default plotting window, and returns the

Immerse.file_backends

Immerse.find_inmask

Immerse.find_object

Immerse.find_panelforms

Immerse.find_panelforms!

Immerse.find_parent

Immerse.find_path

Immerse.find_path!

Immerse.find_tagged

Immerse.find_tagged!

Immerse.fullview_cb

Immerse.fullview_wrapper

Immerse.gcf — Function.

gcf() ("get current figure") returns the current figure number

Immerse.getfig

Immerse.getproperty

Immerse.gridstack

Immerse.gtkdestroy

Immerse.gtkwindow

Immerse.has_tag

Immerse.hasfig

Immerse.hexagon

Immerse.hit — Function.

hit((fig,tag), cb) turns on hit-testing for the plot element tagged with tag in figure fig. When the user clicks on the element, the callback function cb will be executed.

The callback should have the form

    function my_callback(figtag, index, event, distance)
        if distance < 2
            # We clicked close enough to the object to "count"
            # implement the action here
        end
    end

figtag = (figno,tag) identifies the figure and object, and index represents the specific item clicked on. index might be a single Int (e.g., the 37th point in the plot), or for compound objects might be a more complex object. event contains Gtk's information about the click event; this records the x,y position as well as any modifier keys. distance measures how much the user "missed" the object, and is measured in screen-pixels.

Here are a couple of examples:

    fig = display(plot(x=1:10,y=rand(10),Geom.point(tag=:dots)))
    hit((fig,:dots), (figtag, index, xy, dist) -> if dist < 1 println("You clicked on dot ", index) end)

In this case, index will be an Int.

    x = rand(6)
    y = rand(6)
    label = [1,1,1,1,2,2]
    df = DataFrame(Any[x,y,label], [:x,:y,:label])
    fig = display(plot(df, x=:x, y=:y, color=:label, Geom.line(tag=:lines)))

    hit((fig,:lines), (figtag, index, xy, dist) -> begin
        if dist < 2
            println("You clicked on line ", index[1], " in segment ", index[2], " at $(round(Int,100*index[3]))% along the segment")
        end
    end)

By default, hit uses mouse button 1; you can customize the action with hit(figtag, cb, action) where the default is action = :button1press. The choices correspond to the fields in Gtk's MouseHandler.

You can suspend hit-testing for an object with hit((fig,:tag), false) and re-enable it with hit((fig,:tag), true).

Immerse.hitcb

Immerse.hitcenter

Immerse.hline

Immerse.hstack

Immerse.inc

Immerse.inch

Immerse.include

Immerse.initialize_toolbar_callbacks

Immerse.initialized

Immerse.invert_transform

Immerse.iterable

Immerse.lasso_default

Immerse.lasso_initialize — Function.

lasso_initialize(figno, cb) sets the callback cb to run when the user has selected points with the lasso. When you have only one "object" (Geom) in your plot, your callback can be of the form

    function my_simple_lasso_callback(figno, selections)
        println("You selected points ", first(values(selections)))
    end

For a Geom.point object, this would print a vector of integers corresponding to the selected points.

More generally, the callback should have the syntax

    function my_lasso_callback(figno, selections)
        for (form, indexes) in selections
            if form.tag == :mydots
                println("In figure $figno, from :mydots you selected ", indexes)
            end
        end
    end

selections is a Dict of form=>indexes pairs. form is a Compose.Form, the raw objects rendered by Gadfly; you may especially want to query its tag to determine its identity (assuming you've assigned a tag). indexes is a vector describing the items selected; for Circle forms, each element will be an Int, whereas for Line Forms (which can hold multiple lines, perhaps drawn in different colors) each element will be an Tuple{Int,Int} describing the line number and vertex number.

To assign a tag to a Gadfly object, add the tag keyword argument to the geometry, e.g. Geom.point(tag = :mydots).

Immerse.lasso_select_cb

Immerse.lasso_wrapper

Immerse.layer

Immerse.mm

Immerse.native

Immerse.nearest

Immerse.next

Immerse.nextfig

Immerse.no_tweaks

Immerse.octogon

Immerse.on_figure_destroy

Immerse.panzoom_cb

Immerse.panzoom_wrapper

Immerse.path2mask

Immerse.pathx

Immerse.pathy

Immerse.plot

Immerse.ppmm

Immerse.prepare_display

Immerse.proptype2sym

Immerse.pt

Immerse.px

Immerse.render

Immerse.render_backend

Immerse.render_finish

Immerse.save_as

Immerse.save_as_wrapper

Immerse.scf — Function.

scf() ("show current figure") raises (makes visible) the current figure

Immerse.select_points

Immerse.set_default_plot_format

Immerse.set_default_plot_size

Immerse.set_limits!

Immerse.set_ticks!

Immerse.setcoord!

Immerse.setproperty!

Immerse.spy

Immerse.square

Immerse.star1

Immerse.star2

Immerse.start

Immerse.string_compact

Immerse.style

Immerse.switchfig

Immerse.sym2proptype

Immerse.title

Immerse.unblock

Immerse.utriangle

Immerse.vline

Immerse.vstack

Immerse.xcross