Docstrings

SurrealNumbers.*

SurrealNumbers.+

SurrealNumbers.-

SurrealNumbers.<

SurrealNumbers.<=

SurrealNumbers.==

SurrealNumbers.Count

SurrealNumbers.ExistingCanonicals

SurrealNumbers.ExistingConversions

SurrealNumbers.ExistingNegations

SurrealNumbers.ExistingProducts

SurrealNumbers.ExistingSums

SurrealNumbers.ExistingSurreals

SurrealNumbers.Surreal

SurrealNumbers.SurrealDAGstats

SurrealNumbers.SurrealDyadic

SurrealNumbers.SurrealFinite

SurrealNumbers.SurrealMinusOne

SurrealNumbers.SurrealMinusTwo

SurrealNumbers.SurrealNumbers

SurrealNumbers.SurrealOne

SurrealNumbers.SurrealShort

SurrealNumbers.SurrealThree

SurrealNumbers.SurrealTwo

SurrealNumbers.SurrealZero

SurrealNumbers.breadth

SurrealNumbers.canonicalise — Function.

canonicalise(s::SurrealFinite)

Convert a surreal number form into its equivalent canonical form. It performs the conversion by converting to a Rational and then back to a Surreal.

Examples

julia> convert(SurrealFinite, 1) - convert(SurrealFinite, 1)
{ { ϕ | { ϕ | ϕ } } | { { ϕ | ϕ } | ϕ } }
julia> pf( canonicalise( convert(SurrealFinite, 1) - convert(SurrealFinite, 1) ) )
{ ϕ | ϕ }

SurrealNumbers.ceil

SurrealNumbers.clearcache

SurrealNumbers.convert

SurrealNumbers.dag_add

SurrealNumbers.dag_stats

SurrealNumbers.dali

SurrealNumbers.edges

SurrealNumbers.equals

SurrealNumbers.eval

SurrealNumbers.expand — Function.

expand(x::SurrealFinite; level=0)

Writes a surreal as a string with varying levels of expansion.

Arguments

x::SurrealFinite: the number of elements to expand

level=0: the amount of expansion

0 : write shorthand if it exists, or $\{ X_L \| X_R \}$ if not

1 : $\{ X_L \| X_R \}$

2 : expand out $X_L$ and $X_R$ recursively

Examples

julia> expand( convert(SurrealFinite, 2))
"2"
julia> expand( convert(SurrealFinite, 2); level=1)
"{ 1 | ∅ }"
julia> expand( convert(SurrealFinite, 2); level=2)
"{ { { ∅ | ∅ } | ∅ } | ∅ }"

SurrealNumbers.floor

SurrealNumbers.generation — Function.

generation(x::SurrealFinite)

Finds the birthday of a surreal number, which is 1 + the max of any of its components.

Arguments

x::SurrealFinite: the number to operate on

Examples

julia> generation( convert(SurrealFinite, 1) )
1

SurrealNumbers.hash

SurrealNumbers.include

SurrealNumbers.increment

SurrealNumbers.isancestor

SurrealNumbers.iscanonical

SurrealNumbers.isdivisible

SurrealNumbers.iseven

SurrealNumbers.isfinite

SurrealNumbers.isinf

SurrealNumbers.isinteger

SurrealNumbers.isnan

SurrealNumbers.isodd

SurrealNumbers.leq

SurrealNumbers.mod

SurrealNumbers.nodes

SurrealNumbers.one

SurrealNumbers.parents

SurrealNumbers.paths

SurrealNumbers.pf

SurrealNumbers.pff

SurrealNumbers.promote

SurrealNumbers.read

SurrealNumbers.reset!

SurrealNumbers.round

SurrealNumbers.show

SurrealNumbers.sign

SurrealNumbers.spf

SurrealNumbers.surreal2dag — Function.

surreal2dag(x::SurrealFinite)
surreal2dag(io::IO, x::SurrealFinite)

Writes a surreal representation as a DAG out in DOT format for drawing using GraphVis, and returns the number of nodes in the graph. Returns the number of nodes in the graph.

Arguments

io::IO: output stream, default is stdout

x::SurrealFinite: the number to write out

Examples

julia> surreal2dag(convert(SurrealFinite, 0))
digraph "0.0" {
   node_1 [shape=none,margin=0,label=
         <<TABLE BORDER="0" CELLBORDER="1" CELLSPACING="0" CELLPADDING="4">
         <TR><TD COLSPAN="2">0</TD></TR>
         <TR><TD PORT="L"> ∅ </TD><TD PORT="R"> ∅ </TD></TR>
         </TABLE>>,
         ]; 
}
1

SurrealNumbers.surreal2dag_f

SurrealNumbers.surreal2dot — Function.

surreal2dot(x::SurrealFinite)
surreal2dot(io::IO, x::SurrealFinite)

Writes a surreal representation as a tree out in DOT format for drawing using GraphVis, and returns the number of nodes in the graph.

Arguments

io::IO: output stream, default is stdout

x::SurrealFinite: the number to write out

Examples

julia> surreal2dot(convert(SurrealFinite, 1))
digraph "1.0" {
   node_1 [shape=none,margin=0,label=
         <<TABLE BORDER="0" CELLBORDER="1" CELLSPACING="0" CELLPADDING="4">
         <TR><TD COLSPAN="2">1</TD></TR>
         <TR><TD PORT="L"> <TABLE BORDER="0" CELLBORDER="0" CELLPADDING="0"><TR><TD PORT="0,1"> 0 </TD> &nbsp; </TR></TABLE> </TD><TD PORT="R"> ϕ </TD></TR>
         </TABLE>>,
         ];
   node_1:"0,1" -> node_2;
   node_2 [shape=none,margin=0,label=<<B>0</B>>]
}
2

SurrealNumbers.surreal2dot_f

SurrealNumbers.surreal2node

SurrealNumbers.surreal2tex

SurrealNumbers.surrealDAG

SurrealNumbers.tree_nodes

SurrealNumbers.trunc

SurrealNumbers.unique2!

SurrealNumbers.uniqueness

SurrealNumbers.uniqueness_failure

SurrealNumbers.uniqueness_max

SurrealNumbers.width

SurrealNumbers.zero

SurrealNumbers.ϕ

SurrealNumbers.↑

SurrealNumbers.↓

SurrealNumbers.∅

SurrealNumbers.≀

SurrealNumbers.≅

SurrealNumbers.≇

SurrealNumbers.≺

SurrealNumbers.≻

SurrealNumbers.⪯

SurrealNumbers.⪰