Docstrings

TensorCast.@assertsize — Macro.

@assertsize cond str

Like @assert, but prints both the given string and the condition.

TensorCast.@cast — Macro.

@cast Z[i,j,...] := f(A[j,k,...])  options

Macro for broadcasting, reshaping, and slicing of arrays in index notation. Understands the following things:

A[i,j,k] is a three-tensor with these indices.

B[(i,j),k] is the same thing, reshaped to a matrix. Its first axis (the bracket) is indexed by n = i + (j-1) * N where i ∈ 1:N. This may also be written B[i\j,k].

C[k][i,j] is a vector of matrices.

D[j,k]{i} is an ordinary matrix of SVectors, which may be reinterpreted from A[i,j,k].

E[i,_,k] has two nontrivial dimensions, and size(E,2)==1. On the right hand side (or when writing to an existing array) you may also write E[i,3,k] meaning view(E, :,3,:), or E[i,$c,j] to use a variable c. Fixing inner indices, like C[k][i,_], is not allowed.

F[i,-j,k] means reverse(F, dims=2).

The left and right hand sides must have all the same indices. See @reduce for a related macro which can sum over things.

If several tensors appear on the right hand side, then this represents a broadcasting operation, and the necessary re-orientations of axes are automatically inserted.

The following actions are possible:

= writes into an existing array.

:= creates a new object... which may or may not be a view of the input:

== insists on a view of the old object (error if impossible), and |= insists on a copy.

Options can be specified at the end (if several, separated by , i.e. options::Tuple)

i:3 supplies the range of index i. Variables j:rangej and functions k:length(K) are allowed.

assert or ! will turn on explicit dimension checks.

cat will glue slices by things like hcat(A...) instead of reduce(hcat, A), and lazy will instead make a VectorOfArrays container.

strided will place @strided in front of broadcasting operations, and use @strided permutedims(A, ...) instead of PermutedDimsArray(A, ...).

Static slices D[j,k]{i} need using StaticArrays, and to create them you should give all slice dimensions explicitly. You may write D[k]{i:2,j:2} to specify Size(2,2) slices.

TensorCast.@cast! — Macro.

@cast! Z[i...] := A[j...] opt

Variant of @cast which effectively runs @check!() on each tensor.

TensorCast.@check! — Macro.

@check!(A[i, j, μ, ν])

Adds A to the store of known tensors, and records that it expects indices i,j,μ,ν. If it is already in the store, then instead this checks whether the present indices differ from the saved ones. Only the first letter is examined: α and α2 are similar, as are nearby letters β, γ3. More complicated indices like Z[(i,j), -k, _, 3] will be ignored. This happens while parsing your source code, there is zero run-time penalty. Returns A.

In addition, with size=true option, it can insert size checks to be performed at run-time, by returning check!(A, stuff). At the first occurrance this saves i: => size(A,1) etc., and on subsequent uses of the same index (even on different tensors) it gives an error if the sizes do not match. Here the whole index is used: α, β and β2 may have different ranges. This will need to look up indices in a dictionary, which takes ... 50ns, really?

@check! B[i,j] C[μ,ν] D[i] E[j]

Checks several tensors, returns nothing.

@check!  alpha=true  tol=3  size=false  throw=false  info  empty

Controls options for @check! and related macros (@shape!, @reduce!, @einsum! etc). These are the default settings:

alpha=true turns on the parse-time checking, based on index letters.

tol=3 sets how close together letters must be: B[j,k] is not an error but B[a,b] will be.

size=false turns off run-time size checking.

throw=false means that errors are given using @error, without interrupting your program.

empty deletes all saved letters and sizes – there is one global store for each, for now.

info prints what's currently saved.

@cast! B[i,j] := D[i] * E[j]
@reduce! B[i,j] := sum(μ,ν) A[i,j,μ,ν] / C[μ,ν]

@einsum!  B[i,j] := D[i] * E[j]
@vielsum! B[i,j] := D[i] * E[j]
@tensor!  B[i,j] := D[i] * E[j]

Versions of the macros from this package, and from Einsum.jl and TensorOperations.jl, which call @check! on each of their tensors, before proceeding as normal.

TensorCast.@einsum! — Macro.

@einsum! A[i,j] := B[i,k] * C[k,j]

Variant of @einsum from package Einsum.jl, equivalent to wrapping every tensor with @check!().

TensorCast.@pretty — Macro.

@pretty @shape A[...] := B[...]

Prints an approximately equivalent expression with the macro expanded. Compared to @macroexpand1, generated symbols are replaced with animal names, comments are deleted, module names are removed from functions, and the final expression is fed to println().

TensorCast.@reduce — Macro.

@reduce A[i] := sum(j,k) B[i,j,k]             # A = vec(sum(B, dims=(2,3))) 
@reduce A[i] := prod(j) B[i] + ε * C[i,j]     # A = vec(prod(B .+ ε .* C, dims=2))
@reduce A[i] = sum(j) exp( C[i,j] / D[j] )    # sum!(A, exp.(C ./ D') )

Tensor reduction macro:

The reduction function can be anything which works like sum(B, dims=(1,3)), for instance prod and maximum and Statistics.mean.

In-place operations Z[j] = sum(... will construct the banged version of the given function's name, which must work like sum!(Z, A).

The tensors can be anything that @cast understands, including gluing of slices B[i,k][j] and reshaping B[i\j,k].

Index ranges may be given afterwards (as for @cast) or inside the reduction sum(i:3, k:4).

All indices appearing on the right must appear either within sum(...) etc, or on the left.

F = @reduce sum(i,j)  B[i] + γ * D[j]         # sum(B .+ γ .* D')
@reduce G[] := sum(i,j)  B[i] + γ * D[j]      # F == G[]

Complete reduction to a scalar output F, or a zero-dim array G.

@reduce Z[k] := sum(i,j) A[i] * B[j] * C[k]  lazy, i:N, j:N, k:N

The option lazy replaces the broadcast expression with a BroadcastArray, to avoid materializeing the entire array (here size N^3) before summing.

The option strided will place @strided in front of the broadcasting operation. You need using Strided for this to work.

TensorCast.@reduce! — Macro.

@reduce! Z[j] := sum(i,k) A[i,j,k]

Variant of @reduce which effectively runs @check!() on each tensor.

TensorCast.@tensor! — Macro.

@tensor! A[i,j] := B[i,k] * C[k,j]

Variant of @tensor from package TensorOperations.jl, equivalent to wrapping every tensor with @check!().

TensorCast.@vielsum! — Macro.

@vielsum! A[i,j] := B[i,k] * C[k,j]

Variant of @vielsum from package Einsum.jl, equivalent to wrapping every tensor with @check!().

TensorCast.CheckOptions

TensorCast.Reverse

TensorCast.Shuffle

TensorCast.SizeDict

TensorCast.StaticArrays

TensorCast.TensorCast

TensorCast.V

TensorCast.__init__

TensorCast._check!

TensorCast._einsum!

TensorCast._macro

TensorCast._tensor!

TensorCast.cat_glue

TensorCast.check! — Function.

check!(A, (:i,:j), "A[i,j]", (mod=..., src=...))

Performs run-time size checking, on behalf of the @check! macro, returns A. The string and tuple are just for the error message.

TensorCast.check_err

TensorCast.check_one — Function.

check_one(A[i,j,k], (mod=Module, src=...))

Does the work of @check!, on one index expression, returning A or check!(A,...) according to global flags.

TensorCast.check_options

TensorCast.checkrepeats — Function.

checkrepeats(flat)

Throws an error if there are repeated indices.

TensorCast.colonise! — Function.

colonise!(isz, slist)

This aims to simplify "isz" which is going to be used reshape(A, isz). Partly for cosmetic reasons... some of which are now caught elsewhere.

But partly because slist may contain one :, and isz may try to multiply that by sz[d], correct answer is : again.

TensorCast.copy_glue

TensorCast.countcolons

TensorCast.decolonise

TensorCast.eval

TensorCast.findcheck

TensorCast.flag_list

TensorCast.glue — Function.

glue!(B, A, code)
glue(A, code) = glue!(Array{T}(...), A, code)

Copy the contents of an array of arrays into one larger array, un-doing sliceview / slicecopy with the same code. Also called stack or align elsewhere.

cat_glue(A, code)
red_glue(A, code)

The same result, but calling either things like hcat(A...) or things like reduce(hcat, A). The code must be sorted like (:,:,:,*,*), except that (*,:) is allowed.

TensorCast.glue! — Function.

glue!(B, A, code)
glue(A, code) = glue!(Array{T}(...), A, code)

Copy the contents of an array of arrays into one larger array, un-doing sliceview / slicecopy with the same code. Also called stack or align elsewhere.

cat_glue(A, code)
red_glue(A, code)

The same result, but calling either things like hcat(A...) or things like reduce(hcat, A). The code must be sorted like (:,:,:,*,*), except that (*,:) is allowed.

TensorCast.gluecodecheck

TensorCast.gluedsize

TensorCast.include

TensorCast.index_store

TensorCast.inputex — Function.

inputex(:( A[i,j][k] ), canon, flags, store, icheck, where)

Figures out all the steps needed to transform the given tensor to a boring one, aligned with canonical, and returns the necessary expression. Write sizes which can be read from A into store, and necessary reshapes in terms of sz[d].

TensorCast.iscodesorted

TensorCast.isconstant

TensorCast.makelazy — Function.

makelazy(bc)

Takes the result of Broadcast.__dot__() and converts it to have a LazyArrays.BroadcastArray.

TensorCast.oddunique — Function.

oddunique(negated)

Returns a list in which anything repeated evenly many times has been removed, then unique.

TensorCast.orient — Function.

B = orient(A, code)

Reshapes A such that its nontrivial axes lie in the directions where code contains a :, by inserting axes on which size(B, d) == 1 as needed.

TensorCast.outputinplace — Function.

outputinplace(newright, outUZ, redfun, canonsize, canon, flags, store, nameZ)

For the case of = this figures out how to write RHS into LHS, in one of three ways:

reduction sum!(Z, newright)

broadcasting @. Z[...] = newright

neither, copyto!(Z, newright)

No longer attempts to write permutedims!(Z, A, ...), now just copyto!(Z, PermutedDimsArray(A, ...)). Doesn't really need so many arguments...

TensorCast.outputnew — Function.

 outputnew(newright, outUZ, redfun, canonsize, canon, flags, store)

For the case of :=, this constructs the expression to do reduction if needed, and slicing/reshaping/reversing for LHS.

outUZ = (redUind, negV, codeW, indW, sizeX, getY, numY, sizeZ)

TensorCast.packagecheck

TensorCast.packageerror

TensorCast.parse! — Function.

flat, getafix, putsz, negated = parse!(sdict, A, outer, inner)

Use this for A[outer...][inner...].

flat is a list of symbols, naming the indices, in order.

negated is the subset which had - in front.

getafix is the thing for making views of given input or target array, : in non-fixed directions.

putsz is a tuple of sz[n] and products, which you want for reshaping to the very final array; the numbers n refer to position in flat.

sdict::SizeDict ends up with sdict.dict[:i] = an expr for length of this index. To avoid saving these (if you don't have A yet) set A = nothing.

parse!(sdict, A, outer, reduce_inner, true)

This is for A[outer...] = sum(inner...) ... LHS, which are allowed to have sum(a:2, ...) ranges. The allowranges flag also disables the use of size(first(A), 2) etc, to disable size(A,2) set A = nothing again.

parse!(sdict, nothing, [], rex, true, flag_vector)

For dimensions & annotations. Will look for flags because it was given something to push them into. You should now do this first, so that sdict is most likely to have these (neater) entries. They are added using savesize!(sdict,...) which now puts later sizes into list of checks.

TensorCast.pretty

TensorCast.push_or_append!

TensorCast.readleft — Function.

canon, outUZ, nameZ, checkZ = readleft(left, redind, flags, store, icheck, where)

outUZ = (redUind, negV, codeW, sizeX, getY, numY, sizeZ) are things passed to output construction.

TensorCast.recursive_glue

TensorCast.red_glue

TensorCast.savesize!

TensorCast.simplicode — Function.

p, (bef, aft) = simplicode(p, (bef, aft))

The idea here is to absorb permutedims into alterations of the slicing / gluing codes. Done as an optimisation pass, after figuring all of those out, before building expression.

If we also have reversal or shift, those dimensions will need to be updated. Maybe then don't bother.

If there is a reduction, then it's totally worth bothering! You have more freedom than with slicing.

It will also tend to break agreement between canonical sizes & size of actual array at any stage. This may matter for in-place case.

For now, only apply this when those are not in use.

TensorCast.size_store

TensorCast.sizeinfer — Function.

sizeinfer(store, icannon, leaveone=true)

This is the point of SizeDict. The goal is to produce a canonical vector of sizes, corresponding to vector of symbols icannon. If these sizes are known, easy!

But for unknown ones, we do a second pass, looking for entries in sizedict like [:i, :j] which come from tuples of indices, for which we know the product of their dimensions.

TensorCast.slicecopy — Function.

sliceview(A, code)
slicecopy(A, code)

Slice array A according to code, a tuple of length ndims(A), in which : indicates a dimension of the slices, and * a dimension separating them. For example if code = (:,*,:) then slices are either view(A, :,i,:) or A[:,i,:] with i=1:size(A,2).

TensorCast.sliceview — Function.

sliceview(A, code)
slicecopy(A, code)

Slice array A according to code, a tuple of length ndims(A), in which : indicates a dimension of the slices, and * a dimension separating them. For example if code = (:,*,:) then slices are either view(A, :,i,:) or A[:,i,:] with i=1:size(A,2).

TensorCast.sorted_starcode

TensorCast.static_glue

TensorCast.static_slice

TensorCast.stripminus! — Function.

stripminus(negated, i)

For any index, or vector of indices, or vector of indices containing tuples / backslash combos, this pushes every letter with a minus in front into negated list.

It returns an index, or a tidied-up vector of indices, in which tuples etc have become vectors.

TensorCast.szwrap

TensorCast.walker — Function.

walker(outex, x, canon, flags, store, icheck, where)

Called by MacroTools.prewalk on RHS, finds tensors & pushes :( sym = inputex(this) ) into outex.args, and then replaces them with sym.