Docstrings

WeightedArrays.@pack_WeightOpt

WeightedArrays.@pack_WeightOpt!

WeightedArrays.@unpack_WeightOpt

WeightedArrays.ALPHA

WeightedArrays.AbsArray

WeightedArrays.AbsMat

WeightedArrays.AbsVec

CatViews.CatView — Type.

CatView(x::Weighted)

Gives a vector-like view of both x.array and x.weights, the last d components are weights.

WeightedArrays.ClampedWeighted

WeightedArrays.DIGITS

WeightedArrays.GroupSlices

WeightedArrays.MINPROB

WeightedArrays.MINWEIGHT

WeightedArrays.MaybeWeightedArray

WeightedArrays.MaybeWeightedMatrix

WeightedArrays.MaybeWeightedVector

WeightedArrays.PLOTSIZE

WeightedArrays.SAVED_PCA

WeightedArrays.UnClampedWeighted

WeightedArrays.WeightOpt

WeightedArrays.Weighted — Type.

x = Weighted(array, weights)

Created a Weighted Array, with given weights associated to the last dimension of array, e.g. to columns of a matrix. Keyword norm=true specifies that weights(x) sum to 1 always; this is enforced on construction. Calling normalise!(x) will re-enforce this, after scaling 2x or hcat()-ing x + y or mutating.

x = Weighted(array, weights, lo, hi)

Here every θ ∈ array is constrained lo ≦ θ ≦ hi, the same for all dimensions. (Weights are similarly 0 ≦ λ < ∞.) Calling clamp!(x) will re-enforce these box constraints, for instance after mutating x[1,2] *= 3 or on y = x .+ 0.4.

WeightedArrays.WeightedArray

WeightedArrays.WeightedArrays

WeightedArrays.WeightedMatrix

WeightedArrays.WeightedVector

WeightedArrays.__init__

WeightedArrays.addlname

WeightedArrays.addname

WeightedArrays.aname

WeightedArrays.array

WeightedArrays.clampdoc

WeightedArrays.clip — Function.

clip(x, ϵ=1.0e-100)
clip!(x)

Sets to zero all entries abs(x[i]) < ϵ. When x::Weighted this acts on x.array only; see also trim().

WeightedArrays.clip! — Function.

clip(x, ϵ=1.0e-100)
clip!(x)

Sets to zero all entries abs(x[i]) < ϵ. When x::Weighted this acts on x.array only; see also trim().

WeightedArrays.eval

WeightedArrays.flatcopy — Function.

flatten(x::Weighted)

Makes a vector [ reshaped(x.array,:) ; x.weights ].

flatcopy!(x, v::Vector)

Copies numbers from v = flatten(x) back into x.

flatcopy(x, v)

Uses only size(x) to know what shape ::Weighted to make out of v (and x.opt for details).

WeightedArrays.flatcopy! — Function.

flatten(x::Weighted)

Makes a vector [ reshaped(x.array,:) ; x.weights ].

flatcopy!(x, v::Vector)

Copies numbers from v = flatten(x) back into x.

flatcopy(x, v)

Uses only size(x) to know what shape ::Weighted to make out of v (and x.opt for details).

WeightedArrays.flatten — Function.

flatten(x::Weighted)

Makes a vector [ reshaped(x.array,:) ; x.weights ].

flatcopy!(x, v::Vector)

Copies numbers from v = flatten(x) back into x.

flatcopy(x, v)

Uses only size(x) to know what shape ::Weighted to make out of v (and x.opt for details).

WeightedArrays.flipzero

WeightedArrays.grsafe

WeightedArrays.include

WeightedArrays.lastlength

WeightedArrays.load — Function.

save(Π, "file.csv") = writecsv("file.csv", Π)
save(Π, "file.json")
save(Π, "file.jld") = JLD.save(file, "Π", Π)

Save Π::Weighted to disk, in one of several formats:

CSV: weights are rightmost column, as in [Π.array' Π.weights]. Options aren't saved, and load() will try to guess them from the numbers. Mathematica: arr = Import["file.csv"][[;; , 1 ;; -2]]; wei = Import["file.csv"][[;; , -1]]

JSON: saves a dictionary, but dict["array"] is a nested Vector{Vector{Any}} for now, which load() converts to Float64. Includes dict["opt"] == String(Π.opt). Mathematica: arr = Import["file.json"][[1, 2]]; wei = Import["file.json"][[3, 2]] ... and then maybe Style @@@ Thread[{arr, PointSize /@ (0.3*Sqrt[wei]), Opacity[0.5]}] // ListPlot ?

JLD: built-in HDF5 format binary saving, fast & neat but perhaps fragile. Removed for now!

Π = load("file.csv") = readcsv("file.csv", Π)
Π = load("file.json")
Π = load("file.jld") = JLD.load(file, "Π")

Reverse the above.

WeightedArrays.maxcol — Function.

maxcol(x)
maxcol(x, weights)

Returns the column of x with the largest weight, as the same type. If x::Weighted{Matrix} then weights need not be given.

WeightedArrays.maxweight

WeightedArrays.myround

WeightedArrays.near

WeightedArrays.normalise — Function.

normalise(x) ## with an s, NB!
normalise!(x)

Ensures weights are positive and sum to 1. On x::Weighted... the mutating form checks whether norm=true in x.opt; the copying (!) form sets this flag first.

unnormalise(x)
unnormalise!(x)

Just alters the flag x.opt.norm.

WeightedArrays.normalise! — Function.

normalise(x) ## with an s, NB!
normalise!(x)

Ensures weights are positive and sum to 1. On x::Weighted... the mutating form checks whether norm=true in x.opt; the copying (!) form sets this flag first.

unnormalise(x)
unnormalise!(x)

Just alters the flag x.opt.norm.

WeightedArrays.pairwise2 — Function.

pairwise2(x, y=x) = Distances.pairwise(SqEuclidean(), x, y)

Resulting mat[i,j] is distance sqared from x[:,i] to y[:,j]. Implementation varies.

WeightedArrays.pcaylim

WeightedArrays.pointsize

WeightedArrays.positive

WeightedArrays.pplot

WeightedArrays.pplot!

WeightedArrays.rPCA — Function.

f = wPCA(x::Weighted, d=2)
f(x)::Weighted

Works out the MultivariateStats.PCA to project x into d dimensions. To ignore weights, use wPCA(array(x), d) instead. Returned object can be used as a function f(y) ≈ transform(f,array(y)), preserving weights(y).

sPCA(x, d=2)
rPCA(y)

Version with s saves the function above to a global variable, and also applies it, returning f(x). Then you can recall it, and apply it to some other data, with rPCA(y). For example:

julia> xx = wrandn(7,50); yy = xx[:, 1:10] |> normalise
julia> plot(xx, sPCA); plot!(yy, rPCA)

Now equivalent to using PCA-plot function: pplot(xx); pplot!(yy).

WeightedArrays.readcsv

WeightedArrays.readjson

WeightedArrays.rpca — Function.

f = wPCA(x::Weighted, d=2)
f(x)::Weighted

Works out the MultivariateStats.PCA to project x into d dimensions. To ignore weights, use wPCA(array(x), d) instead. Returned object can be used as a function f(y) ≈ transform(f,array(y)), preserving weights(y).

sPCA(x, d=2)
rPCA(y)

Version with s saves the function above to a global variable, and also applies it, returning f(x). Then you can recall it, and apply it to some other data, with rPCA(y). For example:

julia> xx = wrandn(7,50); yy = xx[:, 1:10] |> normalise
julia> plot(xx, sPCA); plot!(yy, rPCA)

Now equivalent to using PCA-plot function: pplot(xx); pplot!(yy).

WeightedArrays.sPCA — Function.

f = wPCA(x::Weighted, d=2)
f(x)::Weighted

Works out the MultivariateStats.PCA to project x into d dimensions. To ignore weights, use wPCA(array(x), d) instead. Returned object can be used as a function f(y) ≈ transform(f,array(y)), preserving weights(y).

sPCA(x, d=2)
rPCA(y)

Version with s saves the function above to a global variable, and also applies it, returning f(x). Then you can recall it, and apply it to some other data, with rPCA(y). For example:

julia> xx = wrandn(7,50); yy = xx[:, 1:10] |> normalise
julia> plot(xx, sPCA); plot!(yy, rPCA)

Now equivalent to using PCA-plot function: pplot(xx); pplot!(yy).

WeightedArrays.save — Function.

save(Π, "file.csv") = writecsv("file.csv", Π)
save(Π, "file.json")
save(Π, "file.jld") = JLD.save(file, "Π", Π)

Save Π::Weighted to disk, in one of several formats:

CSV: weights are rightmost column, as in [Π.array' Π.weights]. Options aren't saved, and load() will try to guess them from the numbers. Mathematica: arr = Import["file.csv"][[;; , 1 ;; -2]]; wei = Import["file.csv"][[;; , -1]]

JSON: saves a dictionary, but dict["array"] is a nested Vector{Vector{Any}} for now, which load() converts to Float64. Includes dict["opt"] == String(Π.opt). Mathematica: arr = Import["file.json"][[1, 2]]; wei = Import["file.json"][[3, 2]] ... and then maybe Style @@@ Thread[{arr, PointSize /@ (0.3*Sqrt[wei]), Opacity[0.5]}] // ListPlot ?

JLD: built-in HDF5 format binary saving, fast & neat but perhaps fragile. Removed for now!

Π = load("file.csv") = readcsv("file.csv", Π)
Π = load("file.json")
Π = load("file.jld") = JLD.load(file, "Π")

Reverse the above.

WeightedArrays.shadowxy

WeightedArrays.sigmoid — Function.

sigmoid(Π)

Compactifies all entries to [0,1], approximately == sigmoid.(Π) but with nice labels etc.

WeightedArrays.sobol — Function.

sobol(d, k)

First k entries from the d-dimensional Sobol sequence, as columns of a Weighted{Matrix}.

WeightedArrays.soboln — Function.

soboln(d, k)

Normally distributed d-vectors, of mean zero and std. dev. π, generated from Sobol sequence, as columns of a Weighted{Matrix}.

WeightedArrays.sobolnp

WeightedArrays.sortcols — Function.

sortcols(x::WeightedMatrix)

Sorts according to θ₁, i.e. the first row of array(x).

WeightedArrays.spca — Function.

f = wPCA(x::Weighted, d=2)
f(x)::Weighted

Works out the MultivariateStats.PCA to project x into d dimensions. To ignore weights, use wPCA(array(x), d) instead. Returned object can be used as a function f(y) ≈ transform(f,array(y)), preserving weights(y).

sPCA(x, d=2)
rPCA(y)

Version with s saves the function above to a global variable, and also applies it, returning f(x). Then you can recall it, and apply it to some other data, with rPCA(y). For example:

julia> xx = wrandn(7,50); yy = xx[:, 1:10] |> normalise
julia> plot(xx, sPCA); plot!(yy, rPCA)

Now equivalent to using PCA-plot function: pplot(xx); pplot!(yy).

WeightedArrays.sureweights

WeightedArrays.svecs

WeightedArrays.totweight

WeightedArrays.trim — Function.

trim(x::Weighted)
trim!(x::Weighted)

Removes points with weight less than cutoff cut=1.0e-10, either copying or mutating.

trim!(Π::Weighted, L::Weighted)

Removes the same columns from both, using the first one's weights. Mutates both arguments! Returns tuple (Π,L).

WeightedArrays.trim! — Function.

trim(x::Weighted)
trim!(x::Weighted)

Removes points with weight less than cutoff cut=1.0e-10, either copying or mutating.

trim!(Π::Weighted, L::Weighted)

Removes the same columns from both, using the first one's weights. Mutates both arguments! Returns tuple (Π,L).

WeightedArrays.trim_

WeightedArrays.unclamp

WeightedArrays.unique_

WeightedArrays.uniquedoc

WeightedArrays.unnormalise

WeightedArrays.unnormalise!

WeightedArrays.wPCA — Function.

f = wPCA(x::Weighted, d=2)
f(x)::Weighted

Works out the MultivariateStats.PCA to project x into d dimensions. To ignore weights, use wPCA(array(x), d) instead. Returned object can be used as a function f(y) ≈ transform(f,array(y)), preserving weights(y).

sPCA(x, d=2)
rPCA(y)

Version with s saves the function above to a global variable, and also applies it, returning f(x). Then you can recall it, and apply it to some other data, with rPCA(y). For example:

julia> xx = wrandn(7,50); yy = xx[:, 1:10] |> normalise
julia> plot(xx, sPCA); plot!(yy, rPCA)

Now equivalent to using PCA-plot function: pplot(xx); pplot!(yy).

WeightedArrays.wcopy

WeightedArrays.wcopy!

WeightedArrays.weights

WeightedArrays.wglue

WeightedArrays.wgrid

WeightedArrays.wname

WeightedArrays.wpca — Function.

f = wPCA(x::Weighted, d=2)
f(x)::Weighted

Works out the MultivariateStats.PCA to project x into d dimensions. To ignore weights, use wPCA(array(x), d) instead. Returned object can be used as a function f(y) ≈ transform(f,array(y)), preserving weights(y).

sPCA(x, d=2)
rPCA(y)

Version with s saves the function above to a global variable, and also applies it, returning f(x). Then you can recall it, and apply it to some other data, with rPCA(y). For example:

julia> xx = wrandn(7,50); yy = xx[:, 1:10] |> normalise
julia> plot(xx, sPCA); plot!(yy, rPCA)

Now equivalent to using PCA-plot function: pplot(xx); pplot!(yy).

WeightedArrays.wrand — Function.

wrand(d, k) = Weighted(rand(d, k))

Uniformly distributed vectors in [0,1]^d, as columns of a Weighted{Matrix} which knows to clamp them to this box. Keyword weights=true gives weights ∝ 1 .+ rand(k) rather than constant. Default is now k=1, making a one-column matrix.

wrandn(d, k) = π .* Weighted(randn(d, k))

Normally distributed d-vectors, of mean zero and std. dev. scale=π by default. Keyword max=10 clamps absolute values to be less than this.

wrandnp(d, k)

Absolute value of normally distributed...

WeightedArrays.wrandn — Function.

wrand(d, k) = Weighted(rand(d, k))

Uniformly distributed vectors in [0,1]^d, as columns of a Weighted{Matrix} which knows to clamp them to this box. Keyword weights=true gives weights ∝ 1 .+ rand(k) rather than constant. Default is now k=1, making a one-column matrix.

wrandn(d, k) = π .* Weighted(randn(d, k))

Normally distributed d-vectors, of mean zero and std. dev. scale=π by default. Keyword max=10 clamps absolute values to be less than this.

wrandnp(d, k)

Absolute value of normally distributed...

WeightedArrays.wrandnp — Function.

wrand(d, k) = Weighted(rand(d, k))

Uniformly distributed vectors in [0,1]^d, as columns of a Weighted{Matrix} which knows to clamp them to this box. Keyword weights=true gives weights ∝ 1 .+ rand(k) rather than constant. Default is now k=1, making a one-column matrix.

wrandn(d, k) = π .* Weighted(randn(d, k))

Normally distributed d-vectors, of mean zero and std. dev. scale=π by default. Keyword max=10 clamps absolute values to be less than this.

wrandnp(d, k)

Absolute value of normally distributed...

WeightedArrays.writecsv

WeightedArrays.writejson

WeightedArrays.wsobol

WeightedArrays.xaxis_grbug

WeightedArrays.xgrid — Function.

xgrid(d, 0:0.1:5)

Gives the matrix whose colums are d-vectors, forming a grid of the given range in all dimensions.

WeightedArrays.yaxis_grbug