Docstrings

GenomicAnnotations.@genes — Macro.

@genes(chr, exs...)

Iterate over and evaluate expressions in exs for all genes in chr.genes, returning genes where all expressions evaluate to true. Any given symbol s in the expression will be substituted for gene.s. The gene itself can be accessed in the expression as gene (see example below).

julia> chromosome = readgbk("example.gbk")
Chromosome 'example' (5028 bp) with 6 annotations

julia> @genes(chromosome, :feature == "CDS") |> length
3

julia> @genes(chromosome, length(gene) < 500)
     CDS             3..206
                     /db_xref="GI:1293614"
                     /locus_tag="tag01"
                     /codon_start="3"
                     /product="TCP1-beta"
                     /protein_id="AAA98665.1"

Accessing properties of the returned list of genes returns a view, which can be altered.

julia> @genes(chromosome, ismissing(:gene)) |> length
2

julia> @genes(chromosome, ismissing(:gene)).gene .= "Unknown";

julia> @genes(chromosome, ismissing(:gene)) |> length
0

GenomicAnnotations.AbstractGene

GenomicAnnotations.Chromosome

GenomicAnnotations.Gene

GenomicAnnotations.GeneDataView

GenomicAnnotations.GenomicAnnotations

GenomicAnnotations.Locus — Type.

Struct for storing information on genomic locations. strand can be '+', '-', or '.' when the strand is irrelevant.

GenomicAnnotations.addgene! — Function.

addgene!(chr::Chromosome, feature, locus; kw...)

Add gene to chr. locus can be a Locus, a UnitRange, or a StepRange.

GenomicAnnotations.appendstring

GenomicAnnotations.eval

GenomicAnnotations.filterseq — Function.

Parse footer (sequence) portion of a GenBank file, returning a String. When BioSequences is available for 0.7 this will be changed to a DNASequence.

GenomicAnnotations.formatsequence

GenomicAnnotations.genes_helper!

GenomicAnnotations.include

GenomicAnnotations.iscomplement — Function.

iscomplement(gene::Abstract)

Return true if gene.locus.compliment == '-', otherwise return false.

GenomicAnnotations.parsechromosome — Function.

parsechromosome(lines)

Parse and return one chromosome entry, and the line number that it ends at.

GenomicAnnotations.parseheader

GenomicAnnotations.parseposition — Function.

Parse lines encoding genomic position, returning the feature as a String, and an instance of Locus.

GenomicAnnotations.printgbk

GenomicAnnotations.pushproperty! — Function.

pushproperty!(gene::AbstractGene, name::Symbol, x::T)

Add a property to gene, similarly to Base.setproperty!(::gene), but if the property is not missing in gene, it will be transformed to store a vector instead of overwriting existing data.

julia> eltype(chr.genedata[:EC_number])
Union{Missing,String}

julia> chr.genes[1].EC_number = "EC:1.2.3.4"
"EC:1.2.3.4"

julia> pushproperty!(chr.genes[1], :EC_number, "EC:4.3.2.1"); chr.genes[1].EC_number
2-element Array{String,1}:
 "EC:1.2.3.4"
 "EC:4.3.2.1"

julia> eltype(chr.genedata[:EC_number])
Union{Missing, Array{String,1}}

GenomicAnnotations.readgbk — Function.

readgbk(filename)

Parse GenBank-formatted file filename, returning a Chromosome.

BioCore.sequence — Function.

sequence(gene::AbstractGene)

Return genomic sequence for gene.

GenomicAnnotations.vectorise — Function.

vectorise(A::AbstractArray{Union{Missing, T}, 1}) where T

Convert an array of type Vector{Union{Missing, T}} to Vector{Union{Missing, Vector{T}}}.