Docstrings

Markovify.Markovify

Markovify.Model — Type.

The datastructure of the Markov chain. Encodes all the different states and the probabilities of going from one to another as a dictionary. The keys are the states, the values are the respective TokenOccurences dictionaries. Those are dictionaries which say how many times was a token found immediately after the state.

Fields

order is the number of tokens in a State

nodes is a dictionary pairing State and its respective

TokenOccurences dictionary.

Markovify.State — Type.

State{T} = Vector{Token{T}}

A state is described by a succession of tokens.

Markovify.Token — Constant.

Token{T} = Union{Symbol, T}

Tokens can be of any type. They can also include symbols :begin and :end which are used to denote the beginning and end of a suptoken.

Markovify.TokenOccurences — Type.

TokenOccurences{T} = Dict{Token{T}, Int}

A dictionary pairing tokens (or special symbols :begin and :end) with the number of their respective occurences.

Markovify.Tokenizer

Markovify.append_token — Function.

append_token(state, token)

Drop the first element in state and append the token at the end of the state array.

Markovify.begseq — Function.

begseq(n)

Return the symbol :begin repeated n times. This array is then used as a starting sequence for all suptokens.

Markovify.combine — Function.

combine(chain, others)

Return a Model which is a combination of all of the models provided. All of the arguments should have the same order. The nodes of all the Models are merged using the function merge.

Markovify.eval

Markovify.include

Markovify.indexin — Function.

indexin(array)

Given a sorted array, return the index on which n would be inserted in should the insertion preserve the sorting.

Markovify.makefromdict

Markovify.next_token — Function.

next_token(model, state)

Return a token which will come after the current state, at random. The probabilities of individual tokens getting choosed are skewed by their individual values in the TokenOccurences dictionary of the current state, that is obtained from the model.

Markovify.randkey — Function.

randkey(dict)

Return a random key from dict. The probabilities of individual keys getting chosen are skewed by their respective values.

Markovify.state_with_beginning

Markovify.state_with_prefix — Function.

state_with_prefix(model, prefix; strict=false)

Attempts to return a random valid state of model that begins with tokens. If strict is false and the model doesn't have any state that begins with tokens, the function shortens the tokens (cuts the last token) to lower the requirements and tries to find some valid state again.

Markovify.states_with_suffix — Function.

states_with_suffix(model, init_suffix)

Return all of the states of model that end with init_suffix. If the number of such states is 1 (or 0), the function shortens the suffix (cuts the first token) in order to lower the requirements, and makes another try.

Markovify.stdweight — Function.

stdweight(state, token)

A constant 1. Used as a placeholder function in Model to represent unbiased weight function.

Markovify.walk — Function.

walk(model[, init_state])

Return an array of tokens obtained by a random walk through the Markov chain. The walk starts at state init_state if supplied, and at state [:begin, :begin...] (the length depends on the order of the supplied model) otherwise. The walk ends once a special token :end is reached.

See also: walk2.

Markovify.walk2 — Function.

walk2(model[, init_state])

Return an array of tokens obtained by a random walk through the Markov chain. When there is only one state following the current one (i.e. there is 100% chance that the state will become the next one), the function shortens the current State as to lower the requirements and obtain more randomness. The State gets shortened until a state with at least two possible successors is found (or until State is only one token long).

The walk starts at state init_state if supplied, and at state [:begin, :begin...] (the length depends on the order of the supplied model) otherwise. The walk ends once a special token :end is reached.

See also: walk.

Markovify.walker — Function.

walker(model, init_state, init_accum, newstate=append_token)

Return an array of tokens obtained by a random walk through the Markov chain. The walk starts at state init_state and ends once a special token :end is reached. A function newstate of general type func(::State{T}, ::Token{T})::State{T} where T can be supplied to be used to generate a new state given the old state and the following token.

This is a general function which is used by all the walk functions.

See also: walk, walk2.