Docstrings

RequiredKeywords.@required_keywords — Macro.

@required_keywords f(...; x::Int=0, y::Int) = ...

Allows function definitions to have keyword arguments without a default value. These keywords must always be specified. If they are not, a UnassignedKeyword exception is thrown.

RequiredKeywords.@showexp

RequiredKeywords.RequiredKeywords — Module.

Implements the @required_keywords macro which allows function definitions to have keyword arguments without a default value. These keywords must always be specified. If they are not, a UnassignedKeyword exception is thrown.

RequiredKeywords.UnassignedKeyword — Type.

Indicated that a funtion has been called without specifying a keyword which has no default.

RequiredKeywords.eval

RequiredKeywords.include