Docstrings

DirectGaussianSimulation.@pack_DirectGaussSimParam

DirectGaussianSimulation.@pack_DirectGaussSimParam!

DirectGaussianSimulation.@unpack_DirectGaussSimParam

DirectGaussianSimulation.DirectGaussSim — Type.

DirectGaussSim(var₁=>param₁, var₂=>param₂, ...)

Direct Gaussian simulation (a.k.a. LU simulation).

Parameters

variogram - theoretical variogram (default to GaussianVariogram())

References

Alabert 1987. The practice of fast conditional simulations through the LU decomposition of the covariance matrix.

DirectGaussianSimulation.DirectGaussSimParam — Type.

DirectGaussSim(var₁=>param₁, var₂=>param₂, ...)

Direct Gaussian simulation (a.k.a. LU simulation).

Parameters

variogram - theoretical variogram (default to GaussianVariogram())

References

Alabert 1987. The practice of fast conditional simulations through the LU decomposition of the covariance matrix.

source

DirectGaussianSimulation.DirectGaussianSimulation

DirectGaussianSimulation.eval

DirectGaussianSimulation.include