Docstrings

SatelliteDynamics.Constants.AS2RAD — Constant.

Constant to convert arcseconds to radians. Equal to 2pi/(360*3600). [rad/as]

SatelliteDynamics.Constants.AU — Constant.

Astronomical Unit. Equal to the mean distance of the Earth from the sun. TDB-compatible value. [m]

P. Gérard and B. Luzum, IERS Technical Note 36, 2010

SatelliteDynamics.Astrodynamics

SatelliteDynamics.Attitude

SatelliteDynamics.Constants.C_LIGHT — Constant.

Speed of light in vacuum. [m/s]

D. Vallado, Fundamentals of Astrodynamics and Applications (4th Ed.), 2010

SatelliteDynamics.Constants

SatelliteDynamics.Coordinates

SatelliteDynamics.Universe.EOP — Constant.

Module-wide global EarthOrientationData object. This data object is used as the default source of Earth Orientation Data by reference system transformations if no explicit EarthOrientationData file is provided to those transformations.

This value can be overridden in your own code as follows:

SatelliteDynamics.EOP = EarthOrientationData(:EOP_PRODUCT_CHOICE)

This global variable defaults to use the module's internal version of :FINALS_2000 if it is not otherwise set/provided.

SatelliteDynamics.Universe.EarthOrientationData — Type.

The EarthOrientationData constains a single data member of type Dict{Int32, Tuple{Float64, Float64, Float64}} that stores the Earth Orientation parameters UT1-UTC, xp, and yp whose units are meters, radians, and radians, respectively. xp and yp are the x- and y-components of Earth's polar motion. The dictionary key is the Epoch the parameters are for as a Modified Julian Day at 0h UTC.

Arguments:

product::Symbol The IERS product type can be :C04_14, :C04_80, or :FINALS_2000

SatelliteDynamics.Time.Epoch — Type.

The Epoch type represents a single instant in time. It is used throughout the SatelliteDynamics module. It is meant to provide a clear definition of moments in time and provide a convenient interface display time in various representations as well as in differrent time systems. The internal data members are also chosen such that the representation maintains nanosecond-precision in reprersenation of time and doesn't accumulate floating-point arithmetic errors larger than nanoseconds even after centuries.

Supports +, +=, -, and -= operators. Two Epoch's can be differenced to return the time difference between two Epochs. If adding a Real number it is interpreted as an offset in seconds to add to the Epoch.

The class also supports all arithmetic operators: ==, !=, <, <=, >, >=

Arguments:

year::Int Year

year::Int Month

year::Int Day

hour::Int Hour (optional)

minute::Int Minute (optional)

second::Real Seconds (optional)

nanoseconds::Real Nanoseconds (optional)

tsys::Symbol: Time system of the epoch at initialization

The Epoch class can be also be initialized from a string. Examples of Valid String constructors are:

epc = Epoch("2018-12-20")
epc = Epoch("2018-12-20T16:22:19.0Z")
epc = Epoch("2018-12-20T16:22:19.123Z")
epc = Epoch("2018-12-20T16:22:19.123456789Z")
epc = Epoch("2018-12-20T16:22:19Z")
epc = Epoch("20181220T162219Z")
epc = Epoch("2018-12-01 16:22:19 GPS")
epc = Epoch("2018-12-01 16:22:19.0 GPS")
epc = Epoch("2018-12-01 16:22:19.123 GPS")
epc = Epoch("2018-12-01 16:22:19.123456789 GPS")

SatelliteDynamics.Attitude.EulerAngle — Type.

The EulerAngle type provides a represenation of EulerAngles for storing attitude information.

Valid sequences are: 121, 123, 131, 132, 212, 213, 231, 232, 312, 313, 321, 323.

Data members:

seq::Symbol: Order of application of angles with respect to body axis.

phi::Float64: First Euler angle

theta::Float64: Second Euler angle

psi::Float64: Third Euler angle

References:

J. Diebel, Representing attitude: Euler angles, unit quaternions, and rotation vectors. Matrix 58(15-16) (2006).

SatelliteDynamics.Attitude.EulerAxis — Type.

The EulerAxis type provides a representation of the Euler angle-and-axis attitude representation.

Data members:

theta::Float64: Angle of rotation

vec::Array{Float64, 1}: Axis of rotation

References:

J. Diebel, Representing attitude: Euler angles, unit quaternions, and rotation vectors. Matrix 58(15-16) (2006).

SatelliteDynamics.Constants.GM_EARTH — Constant.

Earth's Gravitational constant [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_JUPITER — Constant.

Gravitational constant of the Jupiter. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_MARS — Constant.

Gravitational constant of the Mars. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_MERCURY — Constant.

Gravitational constant of the Mercury. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_MOON — Constant.

Gravitational constant of the Moon. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_NEPTUNE — Constant.

Gravitational constant of the Neptune. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_PLUTO — Constant.

Gravitational constant of the Pluto. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_SATURN — Constant.

Gravitational constant of the Saturn. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_SUN — Constant.

Gravitational constant of the Sun. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_URANUS — Constant.

Gravitational constant of the Uranus. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GM_VENUS — Constant.

Gravitational constant of the Venus. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GPS_TAI — Constant.

Offset of GPS time system with respect to TAI time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GPS_TT — Constant.

Offset of GPS time system with respect to TT time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.GPS_ZERO — Constant.

Modified Julian Date of the start of the GPS time system in the GPS time system. This date was January 6, 1980 0H as reckond in the UTC time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Universe.GRAVITY_MODEL — Constant.

Module-wide global GravityModel object. This data object is used as the default spherical harmonic gravity field unless one is otherwise provided.

This value can be overridden in your own code as follows:

SatelliteDynamics.GravityModel = GravityModel(PATH_TO_YOUR_GRAVITY_MODEL)

This global variable defaults to use the module's internal version of the EGM2008 model truncated to order and degree 90, if it is not otherwise set.

SatelliteDynamics.GRAV_COEF

SatelliteDynamics.Universe.GravModel — Type.

GravModel stores a spherical harmonic gravity field in memory. Can store normalized or denomalized coefficients. Package contains EGM2008, GGM01S, and GGM0S gravity models, as well as the default gravity model of EGM2008 truncated to degree and order 90.

Additional gravity field models can be downloaded from: http://icgem.gfz-potsdam.de/home

Arguments:

filepath::string Path to spherical harmonic gravity model file.

SatelliteDynamics.Constants.J2_EARTH — Constant.

Earth's first zonal harmonic. [dimensionless]

GGM05s Gravity Model.

SatelliteDynamics.Constants.MJD2000 — Constant.

Modified Julian Date of January 1, 2000 00:00:00. Value is independent of time scale.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.MJD_ZERO — Constant.

Offset of Modified Julian Days representation with respect to Julian Days. For a time, t, MJD_ZERO is equal to:

MJD_ZERO = t_jd - t_mjd

Where tjd is the epoch represented in Julian Days, and tmjd is the epoch in Modified Julian Days.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.OMEGA_EARTH — Constant.

Earth axial rotation rate. [rad/s]

D. Vallado, Fundamentals of Astrodynamics and Applications (4th Ed.), p. 222, 2010

SatelliteDynamics.OrbitDynamics

SatelliteDynamics.Universe.POLE_LOCATOR — Function.

Compute the location of the pole. Returns x- and y- components as a tuple with the units of [radians]. If the EarthOrientationData argument is ommitted the function will use the default module-global value.

Arguments:

eop::EarthOrientationData EarthOrientationData object to use to compute the offset

mjd::Real Modified Julian Date in UTC of the Epoch for which the pole locator is desired.

interp::Bool Whether to linearly interpolate the parameter data to the input MJD.

Returns:

pole_locator::Tuple{ -Float, Float} (x, y) pole location in radians.

SatelliteDynamics.Constants.P_SUN — Constant.

Nominal solar radiation pressure at 1 AU. [N/m^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Attitude.Quaternion — Type.

The Quaternion type defines a scalar-first quaternion for representation of attitude.

Data members:

q0::Float64: Scalar part of quaternion

q1::Float64: First vector component of quaternion

q2::Float64: Second vector component of quaternion

q3::Float64: Third vector component of quaternion

References:

J. Diebel, Representing attitude: Euler angles, unit quaternions, and rotation vectors. Matrix 58(15-16) (2006).

SatelliteDynamics.Constants.RAD2AS — Constant.

Constant to convert radians to arcseconds. Equal to 2pi/(360*3600) [as/ras]

SatelliteDynamics.Constants.R_EARTH — Constant.

Earth's equatorial radius. [m]

GGM05s Gravity Model

SatelliteDynamics.Constants.R_SUN — Constant.

Nominal solar photospheric radius. [m]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.ReferenceSystems

SatelliteDynamics.Attitude.Rx — Function.

Rotation matrix, for a rotation about the x-axis.

Arguments:

angle::Real: Counter-clockwise angle of rotation as viewed looking back along the postive direction of the rotation axis.

use_degrees:Bool: If true interpret input as being in degrees.

Returns:

r::Array{<:Real, 2}: Rotation matrix

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.27.

SatelliteDynamics.Attitude.Ry — Function.

Rotation matrix, for a rotation about the y-axis.

Arguments:

angle::Real: Counter-clockwise angle of rotation as viewed looking back along the postive direction of the rotation axis.

use_degrees:Bool: If true interpret input as being in degrees.

Returns:

r::Array{<:Real, 2}: Rotation matrix

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.27.

SatelliteDynamics.Attitude.Rz — Function.

Rotation matrix, for a rotation about the z-axis.

Arguments:

angle::Real: Counter-clockwise angle of rotation as viewed looking back along the postive direction of the rotation axis.

use_degrees:Bool: If true interpret input as being in degrees.

Returns:

r::Array{<:Real, 2}: Rotation matrix

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.27.

SatelliteDynamics.SatelliteDynamics

SatelliteDynamics.Simulation

SatelliteDynamics.Constants.TAI_GPS — Constant.

Offset of TAI time system with respect to GPS time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.TAI_TT — Constant.

Offset of TAI time system with respect to TT time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.TT_GPS — Constant.

Offset of TT time system with respect to GPS time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Constants.TT_TAI — Constant.

Offset of TT time system with respect to TAI time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.Time

SatelliteDynamics.Universe.UT1_UTC — Function.

Compute the offset between the UT1 and UTC time systems in seconds. If the EarthOrientationData argument is ommitted the function will use the default module-global value.

Arguments:

eop::EarthOrientationData EarthOrientationData object to use to compute the offset

mjd::Real Modified Julian Date in UTC of the Epoch for which the UT1-UTC offset is desired.

interp::Bool Whether to linearly interpolate the parameter data to the input MJD.

Returns:

ut1_utc::Float UT1 - UTC offset. [s]

SatelliteDynamics.Universe

SatelliteDynamics.Constants.WGS84_a — Constant.

Earth's semi-major axis as defined by the WGS84 geodetic system. [m]

NIMA Technical Report TR8350.2

SatelliteDynamics.Constants.WGS84_f — Constant.

Earth's ellipsoidal flattening. WGS84 Value.

NIMA Technical Report TR8350.2

SatelliteDynamics.Universe.XP — Function.

Compute the x-component of the pole locator in [radians]. If the first EarthOrientationData argument is ommitted the function will use the default module-global value.

Arguments:

eop::EarthOrientationData EarthOrientationData object to use to compute the offset

mjd::Real Modified Julian Date in UTC of the Epoch for which the xp value is desired.

interp::Bool Whether to linearly interpolate the parameter data to the input MJD.

Returns:

xp::Float x-component of pole locator in radians.

SatelliteDynamics.Universe.YP — Function.

Compute the y-component of the pole locator in [radians]. If the first EarthOrientationData argument is ommitted the function will use the default module-global value.

Arguments:

eop::EarthOrientationData EarthOrientationData object to use to compute the offset

mjd::Real Modified Julian Date in UTC of the Epoch for which the yp value is desired.

interp::Bool Whether to linearly interpolate the parameter data to the input MJD.

Returns:

yp::Float y-component of pole locator in radians.

SatelliteDynamics.OrbitDynamics.accel_drag — Function.

Computes the perturbing, non-conservative acceleration caused by atmospheric drag assuming that the ballistic properties of the spacecraft are captured by the coefficient of drag.

Arguments:

x::Array{<:Real, 1}: Satellite Cartesean state in the inertial reference frame [m; m/s]

rho::Real: atmospheric density [kg/m^3]

mass::Real: Spacecraft mass [kg]

area::Real: Wind-facing cross-sectional area [m^2]

Cd::Real: coefficient of drag [dimensionless]

T::Array{<:Real, 2}: Rotation matrix from the inertial to the true-of-date frame

Return:

a::Array{<:Real, 1}: Acceleration due to drag in the X, Y, and Z inertial directions. [m/s^2]

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.83-86.

SatelliteDynamics.OrbitDynamics.accel_gravity — Function.

Computes the accleration caused by Earth gravity as modeled by a spherical harmonic gravity field.

Arguments:

r_sat::Array{<:Real, 1}: Satellite position in the inertial frame [m]

R_eci_ecef::Array{<:Real, 2}: Rotation matrix transforming a vector from the inertial to body-fixed reference frames.

n_max::Integer: Maximum degree coefficient to use in expansion

m_max::Integer: Maximum order coefficient to use in the expansion. Must be less than the degree.

Return:

a::Array{<:Real, 1}: Gravitational acceleration in X, Y, and Z inertial directions [m/s^2]

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.56-68.

SatelliteDynamics.OrbitDynamics.accel_point_mass — Function.

Computes the acceleration of a satellite caused by a point-mass approximation of the central body. Returns the acceleration vector of the satellite.

Assumes the satellite is much, much less massive than the central body.

Arguments:

r_sat::Array{<:Real, 1}: satellite position in a commonn inertial frame [m]

r_body::Array{<:Real, 1}: position of body in a commonn inertial frame [m]

GM::Array{<:Real, 1}: gravitational coeffient of attracting body [m^3/s^2] Default: SatelliteDynamics.Constants.GM_EARTH)

(Default: SatelliteDynamics.Constants.GM_EARTH

Return:

a::Array{<:Real, 1}: Acceleration in X, Y, and Z inertial directions [m/s^2]

Computes the acceleration on a satellite caused by a point-mass approximation of a massive body. Returns the acceleration vector of the satellite.

Arguments:

r_sat::Array{<:Real, 1}: satellite position in the inertial frame [m]

GM::Array{<:Real, 1}: gravitational coeffient of attracting body [m^3/s^2] Default: SatelliteDynamics.Constants.GM_EARTH)

(Default: SatelliteDynamics.Constants.GM_EARTH

Return:

a::Array{<:Real, 1}: Acceleration in X, Y, and Z inertial directions [m/s^2]

SatelliteDynamics.OrbitDynamics.accel_relativity — Function.

Computes perturbation accleration of a satellite in the Inertial frame due to the combined effects of special and general relativity.

Arguments:

x::Array{<:Real, 1}: Satellite Cartesean state in the inertial reference frame [m; m/s]

Returns:

a::Array{<:Real, 1}: Satellite acceleration due to relativity. [m/s^2]

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.110-112.

SatelliteDynamics.OrbitDynamics.accel_srp — Function.

Computes the perturbing acceleration due to direct solar radiation pressure assuming the reflecting surface is a flat plate pointed directly at the Sun.

Arguments:

x::Array{<:Real, 1}: Satellite Cartesean state in the inertial reference frame [m; m/s]

Returns:

a::Array{<:Real, 1}: Satellite acceleration due to solar radiation pressure [m/s^2]

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.77-79.

SatelliteDynamics.OrbitDynamics.accel_thirdbody_moon — Function.

Computes the acceleration of a satellite in the inertial frame due to the gravitational attraction of the Moon.

Arguments:

x::Array{<:Real, 1}: Satellite Cartesean state in the inertial reference frame [m; m/s]

r_moon::Array{<:Real, 1}: Position of moon in inertial frame.

Returns:

a::Array{<:Real, 1}: Acceleration due to the Moon's gravity in the inertial frame [m/s^2]

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.69-70.

SatelliteDynamics.OrbitDynamics.accel_thirdbody_sun — Function.

Computes the acceleration of a satellite in the inertial frame due to the gravitational attraction of the Sun.

Arguments:

x::Array{<:Real, 1}: Satellite Cartesean state in the inertial reference frame [m; m/s]

r_sun::Array{<:Real, 1}: Position of sun in inertial frame.

Return:

a::Array{<:Real, 1}: Acceleration due to the Sun's gravity in the inertial frame [m/s^2]

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.69-70.

SatelliteDynamics.Astrodynamics.anomaly_eccentric_to_mean — Function.

Convert eccentric anomaly into mean anomaly.

Arguments:

E::Real: Eccentric anomaly. [rad] or [deg]

e::Real: Eccentricity. [dimensionless]

use_degrees:Bool: If true interpret input will be interpreted as being in degrees, and output will be returned in degrees.

Returns:

M::Real: Mean anomaly. [rad] or [deg]

SatelliteDynamics.Astrodynamics.anomaly_mean_to_eccentric — Function.

Convert mean anomaly into eccentric anomaly.

Arguments:

M::Real: Mean anomaly. [deg] or [deg]

e::Real: Eccentricity. [dimensionless]

use_degrees:Bool: If true interpret input will be interpreted as being in degrees, and output will be returned in degrees.

Returns:

E::Real: Eccentric anomaly. [rad] or [deg]

SatelliteDynamics.Attitude.as_matrix — Function.

Return the rotation matrix representation of a Quaternion.

Arguments:

q::Quaternion: Quaternion

Returns:

mat::Array{Float64, 2}: Rotation Matrix on SO(3).

SatelliteDynamics.Attitude.as_vector — Function.

Return quaternion as a vector.

Equivalent to q[:]

Arguments:

q::Quaternion: Quaternion

Returns:

vec::Array{Float64, 1}: Quaternion as a (4,) vector

Return Euler angles as a vector.

Equivalent to: [e.phi, e.theta, e.psi] for EulerAngle e

Arguments:

e::EulerAngle Euler Angle

Returns:

evec::Array{Float64, 1} Euler angles components in vector form.

SatelliteDynamics.ReferenceSystems.bias_precession_nutation — Function.

Computes the Bias-Precession-Nutation matrix transforming the GCRS to the CIRS intermediate reference frame. This transformation corrects for the bias, precession, and nutation of Celestial Intermediate Origin (CIO) with respect to inertial space.

Arguments:

epc::Epoch: Epoch of transformation

Returns:

rc2i::Array{<:Real, 2}: 3x3 Rotation matrix transforming GCRS -> CIRS

SatelliteDynamics.Time.caldate — Function.

Return the Gregorian calendar date for a specific

Arguments:

epc::Epoch: Input epoch

tsys::Symbol: Time system to compute output in.

Returns:

year::Int: Year of epoch

month::Int: Month of epoch

day::Int: Day of epoch

hour::Int: Hour of epoch

minute::Int: Minute of epoch

second::Int: Second of epoch

nanoseconds::Int: Year of epoch

SatelliteDynamics.Time.caldate_to_jd — Function.

Convert a Gregorian calendar date to the equivalent Julian Date representation of that time instant.

Aguments:

year::Integer: Year

year::Integer: Month

year::Integer: Day

hour::Integer: Hour

minute::Integer: Minute

second::Real: Seconds

nanoseconds::Real: Nanoseconds

Returns:

mjd::Float64: Julian Date of Epoch

SatelliteDynamics.Time.caldate_to_mjd — Function.

Convert a Gregorian calendar date to the equivalent Modified Julian Date representation of that time instant.

Aguments:

year::Integer Year

year::Integer Month

year::Integer Day

hour::Integer Hour

minute::Integer Minute

second::Real Seconds

nanoseconds::Real Nanoseconds

Returns:

mjd::Float64 Modified Julian Date of Epoch

SatelliteDynamics.Time.day_of_year — Function.

Return the day-of-year number for a given Epoch.

January 1 0h of each year will return 1.

Arguments:

epc::Epoch: Epoch

tsys::Symbol: Time system to return output in

Returns:

doy::Real: Day of year number.

SatelliteDynamics.Time.days_from_elapsed — Function.

Computes the day number in a given time scale given the elapsed time since epoch and the epoch itself.

Assumes all days are counted using a uniform 86400.0 seconds over the time span.

Arguments:

t::Real: Elapsed seconds since the day_epoch.

day_epoch::Real: Day number of the epoch. Common values are SatelliteDynamics.Constants.MJD_ZERO (to get the Julian Day number) or SatelliteDynamics.Constants.MJD2000 (to get Modified Julian Days if reckoning time from January 1, 2000 0H)

Returns:

days::Float: Number of elapsed days in the time scale.

SatelliteDynamics.OrbitDynamics.density_harris_priester — Function.

Computes the local density using the Harris-Priester density model.

Arguments:

x::Array{<:Real, 1}: Satellite Cartesean state in the inertial reference frame [m; m/s]

r_sun::Array{<:Real, 1}: Position of sun in inertial frame.

Returns:

rho:Float64: Local atmospheric density [kg/m^3]

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.89-91.

SatelliteDynamics.OrbitDynamics.deriv_orbit_earth — Function.

Compute the state derivative.

Arguments:

epc::Epoch: Current epoch

x::Array{<:Real, 1}: Satellite state vector

mass::Real: Satellite mass [kg]

area_drag: Velocity-facing area affected by drag. [m^2]

coef_drag: Coefficient of drag [dimensionless]

area_srp: Velocity-facing area affected by drag. [m^2]

coef_srp: Coefficient of reflectivity [dimensionless]

n_grav::Integer: Gravity model degree (Default: 20)

m_grav::Integer: Gravity model order (Default: 20)

drag::Bool: Include cannonball atomospheric drag in force model (Default: true)

srp::Bool: Include flat-plate solar radiation pressure in force model (Default: true)

moon::Bool: Include thridbody lunar gravity in force model (Default: true)

sun::Bool: Include thirdbody solar in force model (Default: true)

relativity::Bool: Include relativistic effects in force model (Default: true)

Returns:

dx::Array{<:Float64, 1}: Satellite state derivative, velocity and accelerations [m; m/s]

SatelliteDynamics.Constants.e_EARTH — Constant.

Earth's first eccentricity. WGS84 Value. [dimensionless]

NIMA Technical Report TR8350.2

SatelliteDynamics.ReferenceSystems.earth_rotation — Function.

Computes the Earth rotation matrix transforming the CIRS to the TIRS intermediate reference frame. This transformation corrects for the Earth rotation.

Arguments:

epc::Epoch: Epoch of transformation

Returns:

r::Array{<:Real, 2}: 3x3 Rotation matrix transforming CIRS -> TIRS

SatelliteDynamics.OrbitDynamics.eclipse_conical — Function.

Computes the illumination fraction of a satellite in Earth orbit using a conical Earth shadow model.

Arguments:

x::Array{<:Real, 1}: Satellite Cartesean state in the inertial reference frame [m; m/s]

r_sun::Array{<:Real, 1}: Position of sun in inertial frame.

Return:

nu::Float64: Illumination fraction (0 <= nu <= 1). nu = 0 means spacecraft in complete shadow, nu = 1 mean spacecraft fully illuminated by sun.

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.80-83.

SatelliteDynamics.OrbitDynamics.eclipse_cylindrical — Function.

Computes the illumination fraction of a satellite in Earth orbit using a cylindrical Earth shadow model.

Arguments:

x::Array{<:Real, 1}: Satellite Cartesean state in the inertial reference frame [m; m/s]

r_sun::Array{<:Real, 1}: Position of sun in inertial frame.

Return:

nu::Float64: Illumination fraction (0 <= nu <= 1). nu = 0 means spacecraft in complete shadow, nu = 1 mean spacecraft fully illuminated by sun.

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.80-83.

SatelliteDynamics.Time.elapsed_from_epoch — Function.

Compute the number of elapsed seconds since a given Epoch from the day number. Can be used to compute the elapsed time since a given Julian or Modified Julian Date.

Arguments:

day_number::Real: Day number, can contain fractional days. Asummes all days are a uniform 86400.0 seconds in length.

day_epoch::Real: Modified Julian Date of Epoch

Returns:

t::Float: Number of elapsed seconds between the Provided Modified Julian date and the epoch.

SatelliteDynamics.Time.epoch_to_jdfd — Function.

Compute the two-part date format used by SOFA.jl functions forr a given Epoch.

Arguments:

epc::Epoch: Epoch

tsys::Symbol: Time system to return output in

Returns:

d1::Real: First part of two part date. [days]

d2::Real: Second part of two part date. [days]

SatelliteDynamics.eval

SatelliteDynamics.Time.gast — Function.

Compute the Greenwich Mean Sidereal Time for the given Epoch.

Arguments:

epc::Epoch: Epoch

use_degrees::Bool: Return output in degrees (Default: false)

Returns:

gast::Real: Greenwich Apparent Sidereal Time [rad/deg]

SatelliteDynamics.Time.gmst — Function.

Compute the Greenwich Mean Sidereal Time for the given Epoch.

Arguments:

epc::Epoch: Epoch

use_degrees::Bool: Return output in degrees (Default: false)

Returns:

gmst::Real: Greenwich Mean Sidereal Time [rad/deg]

SatelliteDynamics.include

SatelliteDynamics.Time.jd — Function.

Compute the Julian Date for a specific epoch

Arguments:

epc::Epoch: Epoch

tsys::Symbol: Time system to return output in

Returns:

jd::Real: Julian date of the epoch in the requested time system

SatelliteDynamics.Time.jd_to_caldate — Function.

Convert a Julian Date to the equivalent Gregorian calendar date representation of the same instant in time.

Aguments:

jd::Real: Julian Date of Epoch

Returns:

year::Int32: Year

year::Int32: Month

year::Int32: Day

hour::Int32: Hour

minute::Int32: Minute

second::Float64: Seconds

microsecond::Float64: Nanoseconds

SatelliteDynamics.Universe.load_eop — Function.

Load new Earth orientation data into the module global EarthOrientationData object. The product can be one of the symbols: :C04_14, :C04_80, or :FINALS_2000.

Arguments:

product::Symbol Loads a different set of EarthOrientationData values into the module-wide global EarthOrientationData parameters.

SatelliteDynamics.Universe.load_gravity_model — Function.

Load new gravity model into module global EarthOrientationData object. The product can be one of the symbols: :EGM2008_20, :EGM2008_90, :GGM01S, :GGM05S, or the filepath to a text-encoded gravity model file.

Arguments:

gfc_file::String File path of gravity field model

product_name::Symbol OR a symbol of a known gravity field product. Valid ones are: :EGM2008_20, :EGM2008_90, :GGM01S, :GGM05S

SatelliteDynamics.Astrodynamics.mean_motion — Function.

Compute the mean motion given a semi-major axis.

Arguments:

a::Real: Semi-major axis. [m]

use_degrees:Bool: If true returns result in units of degrees

GM::Real: Gravitational constant of central body. Defaults to SatelliteDynamics.Constants.GM_EARTH if none is provided.

Returns:

n::Real: Orbital mean motion. [rad/s] or [deg/s]

SatelliteDynamics.Time.mjd — Function.

Compute the Modified Julian Date for a specific epoch

Arguments:

epc::Epoch: Epoch

tsys::Symbol: Time system to return output in

Returns:

mjd::Real: Julian date of the epoch in the requested time system

SatelliteDynamics.Time.mjd_to_caldate — Function.

Convert a Modified Julian Date to the equivalent Gregorian calendar date representation of the same instant in time.

Aguments:

mjd::Real: Modified Julian Date of Epoch

Returns:

year::Int32: Year

year::Int32: Month

year::Int32: Day

hour::Int32: Hour

minute::Int32: Minute

second::Float64: Seconds

nanoseconds::Float64: Nanoseconds

SatelliteDynamics.OrbitDynamics.moon_position — Function.

Compute the Moon's position in the EME2000 inertial frame through the use of low-precision analytical functions.

Argument:

epc::Epoch: Epoch

Returns:

r_moon::Array{<:Real, 1}: Position vector of the Moon in the Earth-centered inertial fame.

Notes:

The EME2000 inertial frame is for most purposes equivalent to the GCRF frame.

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.70-73.

SatelliteDynamics.Astrodynamics.orbit_period — Function.

Compute the satellite orbital period given the semi-major axis.

Arguments:

a::Real: Semi-major axis. [m]

GM::Real: Gravitational constant of central body. Defaults to SatelliteDynamics.Constants.GM_EARTH if none is provided.

Returns:

T::Real: Orbital period. [s]

SatelliteDynamics.ReferenceSystems.polar_motion — Function.

Computes the Earth rotation matrix transforming the TIRS to the ITRF reference frame.

Arguments

epc::Epoch: Epoch of transformation

Returns

rpm::Array{<:Real, 2}: 3x3 Rotation matrix transforming TIRS -> ITRF

SatelliteDynamics.Simulation.propagate_orbit — Function.

Simulate orbit dynamics

Arguments:

epc0::Epoch: Propagation start Epoch

eci0::Epoch: Initial Cartesean inertial state [m; m/s]

epcf::Epoch: Final to simulate to.

timestep::Real: Timestep to use for simulation (Default: 5.0)

solver: Solver to use to solve ODEProblem

atol::Real: Absolute tolerate limit for differential equation solution (Default: 1.0e-9)

atol::Real: Absolute tolerate limit for differential equation solution (Default: 1.0e-9)

mass::Real: Satellite mass (Default: 100.0) [kg]

area_drag::Real: Area of drag (cannon-ball model) (Default: 1.0) [m^2]

coef_drag::Real: Coefficient of drag (Default: 2.3) [dimensionless]

area_srp::Real: Area of solar radiation prerssure (Default: 1.0) [m^2]

coef_srp::Real: Coefficient of reflectivity (Default: 1.8) [dimensionless]

n_grav::Integer: Degree of gravity field (Default: 0.0) [dimensionless]

m_grav::Integer: Order of gravity field (Default: 0.0) [dimensionless]

drag::Bool: Apply drag perturbation (Default: false)

srp::Bool: Apply solar radiation pressure perturbation (Default: false)

moon::Bool: Apply third-body lunar gravity perturbation (Default: false)

sun::Bool: Apply third-body solar gravity perturbation (Default: false)

relativity::Bool: Apply relativistic perturbations (Default: false)

Returns:

t::Array{Float64, 1}: Simulation output times at elapsed seconds from initial Epoch

epc::Array{Epoch, 1}: Simulation output times as absolute Epochs

eci::Array{Float64, 2}: Propgated inertial state. Time is aligned with column dimension.

SatelliteDynamics.ReferenceSystems.rECEFtoECI — Function.

Computes the combined rotation matrix from the Earth-fixed to the inertial reference frame. Applies corrections for bias, precession, nutation, Earth-rotation, and polar motion.

The transformation is accomplished using the IAU 2006/2000A, CIO-based theory using classical angles. The method as described in section 5.5 of the SOFA C transformation cookbook.

Arguments

epc::Epoch: Epoch of transformation

Returns

r::Array{<:Real, 1}: 3x3 Rotation matrix transforming ITRF -> GCRF

SatelliteDynamics.Coordinates.rECEFtoENZ — Function.

Compute the rotation matrix from the Earth-fixed to the East-North-Up coorindate basis.

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Real, 2}: Topocentric rotation matrix

SatelliteDynamics.Coordinates.rECEFtoSEZ — Function.

Compute the rotation matrix from the Earth-fixed to the South-East-Zenith coorindate basis.

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

SatelliteDynamics.ReferenceSystems.rECItoECEF — Function.

Computes the combined rotation matrix from the inertial to the Earth-fixed reference frame. Applies corrections for bias, precession, nutation, Earth-rotation, and polar motion.

The transformation is accomplished using the IAU 2006/2000A, CIO-based theory using classical angles. The method as described in section 5.5 of the SOFA C transformation cookbook.

Arguments

epc::Epoch: Epoch of transformation

Returns

r::Array{<:Real, 2}: 3x3 Rotation matrix transforming GCRF -> ITRF

SatelliteDynamics.ReferenceSystems.rECItoRTN — Function.

Compute the Earth-centered inertial to radial, along-track, cross-track (RTN) rotation matrix. Which, if applied to a position vector in the ECI frame, will transform that vector into the equivalent position vector in the RTN frame.

The RTN frame is also commonly refered to as the local-vertical, local-horizontal (LVLH) frame.

Arguments:

x::Array{<:Real, 1}: Inertial state (position and velocity) of primary (observing) satellite

xt::Array{<:Real, 1}: Inertial state (position and velocity) of the target satellite

Returns:

R_eci_to_rtn::Array{<:Real, 1}: Rotation matrix transforming from the ECI frame to the RTN frame.

SatelliteDynamics.Coordinates.rENZtoECEF — Function.

Compute the rotation matrix from the Earth-fixed to the South-East-Zenith coorindate basis.

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

SatelliteDynamics.ReferenceSystems.rRTNtoECI — Function.

Compute the radial, along-track, cross-track (RTN) rotation matrix. Which, if applied to a position vector in the RTN frame, will transform that vector to beinto the equivalent relative position in the ECI frame.

The RTN frame is also commonly refered to as the local-vertical, local-horizontal (LVLH) frame.

Arguments:

x::Array{<:Real, 1}: Inertial state (position and velocity) of primary (observing) satellite

xt::Array{<:Real, 1}: Inertial state (position and velocity) of the target satellite

Returns:

R_rtn_to_eci::Array{<:Real, 1}: Rotation matrix transforming from the RTN frame to the ECI frame.

SatelliteDynamics.Coordinates.rSEZtoECEF — Function.

Compute the rotation matrix from the Earth-fixed to the South-East-Zenith coorindate basis.

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

SatelliteDynamics.Astrodynamics.sCARTtoOSC — Function.

Given a Cartesean position and velocity in the inertial frame, return the state expressed in terms of osculating orbital elements.

The osculating elements are assumed to be (in order):

a, Semi-major axis [m]

e, Eccentricity [dimensionless]

i, Inclination [rad]

Ω, Right Ascension of the Ascending Node (RAAN) [rad]

ω, Argument of Perigee [ramd]

M, Mean anomaly [rad]

Arguments:

x x::Array{<:Real, 1}: Cartesean inertial state. Returns position and velocity. [m; m/s]

use_degrees:Bool: If true interpret input will be interpreted as being in degrees, and output will be returned in degrees.

GM::Real: Gravitational constant of central body. Defaults to SatelliteDynamics.Constants.GM_EARTH if none is provided.

Returns

x_oe x::Array{<:Real, 1}: Osculating orbital elements. See above for desription of the elements and their required order.

SatelliteDynamics.ReferenceSystems.sECEFtoECI — Function.

Transforms an Earth fixed state into an Inertial state

The transformation is accomplished using the IAU 2006/2000A, CIO-based theory using classical angles. The method as described in section 5.5 of the SOFA C transformation cookbook.

Arguments

epc::Epoch: Epoch of transformation

x::Array{<:Real, 1}: Earth-fixed state (position, velocity) [m; m/s]

Returns

x_ecef::Array{<:Real, 1}: Inertial state (position, velocity)

SatelliteDynamics.Coordinates.sECEFtoENZ — Function.

Compute the coordinates of an object in the topocentric frame of an Earth-fixed frame

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

ecef::Array{<:Real, 1}: Coordinates of the object in Earth-fixed station

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

SatelliteDynamics.Coordinates.sECEFtoGEOC — Function.

Convert Earth-fixed position to geocentric location.

Arguments:

ecef::Array{<:Real, 1}: Earth-fixed coordinated [m]

use_degrees:Bool: If true returns result in units of degrees

Returns:

geoc: Geocentric coordinates (lon, lat, altitude) [rad] / [deg]

SatelliteDynamics.Coordinates.sECEFtoGEOD — Function.

Convert geodetic coordinaties to Earth-fixed position

Arguments:

ecef::Array{<:Real, 1}: Earth-fixed position [m]

use_degrees:Bool: If true returns result in units of degrees

Returns:

geod::Array{<:Real, 1}: Geocentric coordinates (lon, lat, altitude) [rad] / [deg]

SatelliteDynamics.Coordinates.sECEFtoSEZ — Function.

Compute the coordinates of an object in the topocentric frame of an Earth-fixed frame

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

ecef::Array{<:Real, 1}: Coordinates of the object in Earth-fixed station

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

SatelliteDynamics.ReferenceSystems.sECItoECEF — Function.

Transforms an Earth inertial state into an Earth fixed state

The transformation is accomplished using the IAU 2006/2000A, CIO-based theory using classical angles. The method as described in section 5.5 of the SOFA C transformation cookbook.

Arguments

epc::Epoch: Epoch of transformation

x::Array{<:Real, 1}: Inertial state (position, velocity) [m; m/s]

Returns

x_ecef::Array{<:Real, 1}: Earth-fixed state (position, velocity)

SatelliteDynamics.ReferenceSystems.sECItoRTN — Function.

Compute the radial, along-track, cross-track (RTN) coordinates of a target satellite in the primary satellites RTN frame.

The RTN frame is also commonly refered to as the local-vertical, local-horizontal (LVLH) frame.

Arguments:

x::Array{<:Real, 1}: Inertial state (position and velocity) of primary (observing) satellite

xt::Array{<:Real, 1}: Inertial state (position and velocity) of the target satellite

Returns:

rtn::Array{<:Real, 1}: Position and velocity of the target relative of the observing satellite in the RTN.

SatelliteDynamics.Coordinates.sENZtoAZEL — Function.

Convert East-North-Zenith topocentric state to azimuth, elevation, and range.

Arguments:

x::Array{<:Real, 1}: East-North-Up state

use_degrees:Bool: If true returns result in units of degrees

Returns:

azel::Array{<:Real, 1}: Azimuth, elevation and range [rad; rad; m]

SatelliteDynamics.Coordinates.sENZtoECEF — Function.

Compute the coordinates of an object in the topocentric frame of an Earth-fixed frame

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

sez::Array{<:Real, 1}: SEZ coordinates of the object

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

SatelliteDynamics.Coordinates.sGEOCtoECEF — Function.

Convert geocentric position to equivalent Earth-fixed position.

Arguments:

geoc::Array{<:Real, 1}: Geocentric coordinates (lon, lat, altitude) [rad] / [deg]

use_degrees:Bool: If true interpret input as being in degrees.

Returns:

ecef::Array{<:Real, 1}: Earth-fixed coordinates [m]

SatelliteDynamics.Coordinates.sGEODtoECEF — Function.

Convert geodetic position to equivalent Earth-fixed position.

Arguments:

geod::Array{<:Real, 1}: Geodetic coordinates (lon, lat, altitude) [rad] / [deg]

use_degrees:Bool: If true interpret input as being in degrees.

Returns:

ecef::Array{<:Real, 1}: Earth-fixed coordinates [m]

SatelliteDynamics.Astrodynamics.sOSCtoCART — Function.

Given an orbital state expressed in osculating orbital elements compute the equivalent Cartesean position and velocity of the inertial state.

The osculating elements are assumed to be (in order):

a, Semi-major axis [m]

e, Eccentricity [dimensionless]

i, Inclination [rad]

Ω, Right Ascension of the Ascending Node (RAAN) [rad]

ω, Argument of Perigee [ramd]

M, Mean anomaly [rad]

Arguments:

x_oe x::Array{<:Real, 1}: Osculating orbital elements. See above for desription of the elements and their required order.

use_degrees:Bool: If true interpret input will be interpreted as being in degrees, and output will be returned in degrees.

GM::Real: Gravitational constant of central body. Defaults to SatelliteDynamics.Constants.GM_EARTH if none is provided.

Returns

x x::Array{<:Real, 1}: Cartesean inertial state. Returns position and velocity. [m; m/s]

SatelliteDynamics.ReferenceSystems.sRTNtoECI — Function.

Compute the Earth-center

The RTN frame is also commonly refered to as the local-vertical, local-horizontal (LVLH) frame.

Arguments:

x::Array{<:Real, 1}: Inertial state (position and velocity) of primary (observing) satellite

xt::Array{<:Real, 1}: Inertial state (position and velocity) of the target satellite

Returns:

rtn::Array{<:Real, 1}: Position and velocity of the target relative of the observing satellite in the RTN.

SatelliteDynamics.Coordinates.sSEZtoAZEL — Function.

Convert South-East-Zenith topocentric state to azimuth, elevation, and range.

Arguments:

x::Array{<:Real, 1}: South-East-Zenith state

use_degrees:Bool: If true returns result in units of degrees

Returns:

azel::Array{<:Real, 1}: Azimuth, elevation and range [rad; rad; m]

SatelliteDynamics.Coordinates.sSEZtoECEF — Function.

Compute the coordinates of an object in the topocentric frame of an Earth-fixed frame

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

sez::Array{<:Real, 1}: SEZ coordinates of the object

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

SatelliteDynamics.Astrodynamics.semimajor_axis — Function.

Calculate semi-major axis given mean-motion.

Arguments:

n::Real: Orbital mean motion. [rad/s] or [deg/s]

use_degrees:Bool: If true interpret input as being in degrees.

GM::Real: Gravitational constant of central body. Defaults to SatelliteDynamics.Constants.GM_EARTH if none is provided.

Returns:

a::Real: Semi-major axis. [m]

SatelliteDynamics.Universe.set_eop — Function.

Set Earth orientation data values for a specific date in the module global EarthOrientationData object.

Arguments:

mjd::Real Modified Julian Date in UTC of the Epoch for which the Earth orientation data is aligned to.

ut1_utc::Real Offset between UT1 and UTC in seconds.

xp::Real x-component of the pole locator in radians.

yp::Real y-component of the pole locator in radians.

SatelliteDynamics.Attitude.slerp — Function.

Perform spherical linear interpolation (SLERP) on two quaternions. Interpolatles from quaternion, q1, to quaternion, q2, at normalized interpolation time, t.

Interpolation time must be in the range [0, 1] a value of 0 will return q1, while a value of 1 will return q2.

Arguments:

q1::Quaternion: Starting Quaternion

q2::Quaternion: Ending Quaternion

t::Real: Normalized interpolation time. [0, 1]

Returns:

q:Quaternion: Quaternion attitude interpolation from q1 toward q2 at time t.

SatelliteDynamics.OrbitDynamics.sun_position — Function.

Compute the Sun's position in the EME2000 inertial frame through the use of low-precision analytical functions.

Argument:

epc::Epoch: Epoch

Returns:

r_sun::Array{<:Real, 1}: Position vector of the Sun in the Earth-centered inertial fame.

Notes:

The EME2000 inertial frame is for most purposes equivalent to the GCRF frame.

References:

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012, p.70-73.

SatelliteDynamics.Astrodynamics.sun_sync_inclination — Function.

Compute the required inclination for a Sun-synchronous Earth orbit.

Algorithm assumes the nodal precession is entirely due to the J2 perturbation, and no other perturbations are considered.

The inclination is computed using a first-order, non-iterative approximation.

Arguments:

a::Real: Semi-major axis. [m]

e::Real: Eccentricity. [dimensionless]

use_degrees:Bool: If true interpret output will be returned in degrees.

Returns:

iss::Real: Requierd inclination for a sun-synchronous orbit. [rad] or [deg]

SatelliteDynamics.Time.time_system_offset — Function.

Compute the offset between two time systems at a given Epoch.

The offset (in seconds) is computed as:

time_system_offset = tsys_dest - tsys_src

The value returned is the number of seconds that musted be added to the source time system given the input epoch, to get the equivalent epoch.

Conversions are accomplished using SOFA C library calls. Epoch.

Arguments:

jd::Real: Part 1 of two-part date (Julian days)

fd::Real: Part 2 of two-part date (Fractional days)

tsys_src::Symbol: Base time system

tsys_dest::Symbol: Destination time system

Returns:

offset::Float: Offset between soruce and destination time systems in seconds.

SatelliteDynamics.Universe.update_eop — Function.

Download updated Earth orientation datafiles for included products IERS products.

Arguments:

product::Symbol The IERS product type can be :C04_14, :C04_80, or :FINALS_2000