Docstrings

StructuralInheritance.@protostruct — Macro.

@protostruct(struct_ [, prefix_])

creates a struct that can have structure inherited from it if also defined by @protostruct, creates an abstract type with a name given by the struct definitions name and a prefix. The concrete type inherits from the abstract type and anything which inherits the concrrete types structure also inherits from the abstract type.

StructuralInheritance.StructuralInheritance

StructuralInheritance.abstracttype — Function.

Creates an abstract type with the given name

StructuralInheritance.addparams — Function.

attaches parameters to a name

StructuralInheritance.assertcollisionfree — Function.

throws an error is the fields contain overlapping symbols

StructuralInheritance.deparametrize_lightName — Function.

strips parameterization off of a name that does not include inheritence information

StructuralInheritance.eval

StructuralInheritance.extractfields — Function.

extracts the fields from a struct definition

StructuralInheritance.extractname — Function.

gets the name of a struct definition

StructuralInheritance.fieldBacking

StructuralInheritance.fieldsymbols

StructuralInheritance.filtertofields — Function.

returns an array with only the field definitions

StructuralInheritance.fulltypename — Function.

adds source module information to the type name

StructuralInheritance.get2parameters

StructuralInheritance.getparameters

StructuralInheritance.getpath

StructuralInheritance.include

StructuralInheritance.isfunction

StructuralInheritance.isparametric

StructuralInheritance.ispath

StructuralInheritance.newnames

StructuralInheritance.parameterMap

StructuralInheritance.register — Function.

registers a new struct and abstract type pair

StructuralInheritance.rename — Function.

returns a renamed struct

StructuralInheritance.replacefields — Function.

returns a copy with replacement fields

StructuralInheritance.sanitize — Function.

annotates module information to unanotated typed fields

StructuralInheritance.shadowMap

StructuralInheritance.updateParameters — Function.

update parameters from old fields