Docstrings

Suppressor.@capture_err — Macro.

@capture_err expr

Capture the stderr stream for the given expression.

Suppressor.@capture_out — Macro.

@capture_out expr

Capture the stdout stream for the given expression.

Suppressor.@color_output — Macro.

@color_output enabled::Bool expr

Enable or disable color printing for the given expression. Often useful in combination with the @capture_* macros:

Example

@coloroutput false begin output = @captureerr begin @warn "should get captured, not printed" end end @test output == "WARNING: should get captured, not printed "

Suppressor.@suppress — Macro.

@suppress expr

Suppress the stdout and stderr streams for the given expression.

Suppressor.@suppress_err — Macro.

@suppress_err expr

Suppress the stderr stream for the given expression.

Suppressor.@suppress_out — Macro.

@suppress_out expr

Suppress the stdout stream for the given expression.

Suppressor.Suppressor

Suppressor.eval

Suppressor.haslogging

Suppressor.include