Docstrings

DiscreteFunctions.DiscreteFunction — Type.

DiscreteFunction is a function from {1,2,...,n} to itself. It can be created by DiscreteFunction(list) where list is a one-dimensional array of positive integers. Alternatively, it can be created using positive integer arguments.

The following are equivalent:

DiscreteFunction([1,4,2,3])
DiscreteFunction(1,4,2,3)

DiscreteFunctions.DiscreteFunctions

DiscreteFunctions.IdentityFunction — Function.

IdentityFunction(n) creates the identity DiscreteFunction on the set {1,2,...,n}.

DiscreteFunctions.RandomFunction — Function.

RandomFunction(n) creates a random DiscreteFunction on {1,2,...,n}.

DiscreteFunctions.all_functions — Function.

all_functions(n) returns a generator that produces all DiscreteFunctions on {1,2,...,n}

DiscreteFunctions.all_sqrts — Function.

all_sqrts(f::DiscreteFunction) returns an array consisting of all g such that g*g==f.

DiscreteFunctions.all_squares — Function.

all_squares(n) returns a list of all function on [n] that are perfect compositional squares.

DiscreteFunctions.eval

Permutations.fixed_points — Function.

fixed_points(f::DiscreteFunction) returns a list of fixed points of the function, i.e., those values x such that f(x)==x.

DiscreteFunctions.has_inv — Function.

has_inv(f::DiscreteFunction) tests if f is invertible.

DiscreteFunctions.has_sqrt — Function.

has_sqrt(f::DiscreteFunction) checks if there is a function g such that g*g==f. Returns true if so and false otherwise.

DiscreteFunctions.image — Function.

image("f::DiscreteFunction") returns a Set containing all the output values of f.

DiscreteFunctions.include

DiscreteFunctions.is_permutation — Function.

is_permutation(f::DiscreteFunction) returns true if f is a bijection on its domain.

DiscreteFunctions.random_square — Function.

random_square(n) returns a DiscreteFunction on n elements that has a compositional square root.

Base.sqrt — Function.

sqrt(f::DiscreteFunction) returns a DiscreteFunction g such that g*g==f or throws an error if no such g exists.