Docstrings

Rectangle.AbstractBST

Rectangle.AbstractNode

Rectangle.BSTNode

Rectangle.BinarySearchTree

Rectangle.Interval

Rectangle.IntervalKey

Rectangle.IntervalNode

Rectangle.IntervalTree

Rectangle.Iterator

Rectangle.Line

Rectangle.OrderedRectMap

Rectangle.OrderedRectMapX

Rectangle.OrderedRectMapY

Rectangle.RBNode

Rectangle.RBTree

Rectangle.Rect

Rectangle.Rectangle

Rectangle._delete!

Rectangle._delete_fixup!

Rectangle._extremum

Rectangle._inorder

Rectangle._insert!

Rectangle._insert_fixup!

Rectangle._intersect

Rectangle._k

Rectangle._l

Rectangle._l!

Rectangle._maximum

Rectangle._minimum

Rectangle._nv

Rectangle._p

Rectangle._p!

Rectangle._pred_succ

Rectangle._predecessor

Rectangle._r

Rectangle._r!

Rectangle._search

Rectangle._successor

Rectangle._transplant!

Rectangle._v

Rectangle.area

Rectangle.avg_min_dist — Function.

    avg_min_dist(r1::Rect, r2::Rect) -> dx::Float64, dy::Float64

Rectangles are essentially point sets. Hence, one can perceive existence of a minimum distance of one point in r1 from r2. Similar, distance would also exist for every point in r2 from r1. While, technically Euclidean distance metric can exist, the computation is fairly cumborsome. Here, we use the city block distance or L1-metric.

dx: The distance in the x-direction dy: The distance in the y-direction

The minimum distance will be zero when the rectangles are intersecting. The distance also will be lower in a specific direction if there is an overlap of the rectangles in that direction

Rectangle.axis_parallel

Rectangle.c_hi

Rectangle.c_lo

Rectangle.cg

Rectangle.coord

Rectangle.create_ordered_map

Rectangle.delete_rect!

Rectangle.diags

Rectangle.endof

Rectangle.eval

Rectangle.ex

Rectangle.ey

Rectangle.get_intersect_data

Rectangle.get_overlapped_dist

Rectangle.h

Rectangle.has_overlap

Rectangle.has_x_overlap

Rectangle.has_y_overlap

Rectangle.hline_xsection — Function.

Given a Rectangle and a set of a sorted set of horizontal lines ordered
top to bottom, provides the indices that intersect the rectangle.

Rectangle.hlines

Rectangle.horiz_desc — Function.

`isless` function  that can be used to sort horizonal lines in descending
order (top to bottom).

Rectangle.include

Rectangle.insert_rect!

Rectangle.inside

Rectangle.intersect_axis_aligned

Rectangle.intersects — Function.

    intersects(l1::Line{T}, l2::Line{T}) where {T <: Real} -> Bool

If l1 and l2 intersect each other.

Rectangle.isHorizontal — Function.

    isHorizontal(l::Line) -> Bool
    isVertcal(l::Line) -> Bool

If the Line is horizontal or vertical.

Rectangle.isVertical

Rectangle.isnil

Rectangle.lb

Rectangle.left_rotate!

Base.length — Function.

    length(l::Line) -> Float64

The length of the line segment.

Rectangle.line_xsection

Rectangle.lines

Rectangle.lx

Rectangle.ly

Rectangle.merge_axis_aligned — Function.

    merge_axis_aligned(alines::Vector{Line{T}}, 
                       axis::Int=1, 
                       order::Symbol=:increasing,
                       tol::T=pcTol(T)) -> Vector{Line{T}}

Given an array of axis aligned lines, if the line ends touch or have an overlap the lines are merged into a larger segment. Lines which are not touching the other lines are left intact.

order parameter can be in :increasing or :decreasing order in the direction of the axis.

axis parameter can be 1 for horizontal lines and 2 for vertical lines.

Rectangle.min_dist — Function.

    min_dist(r1::Rect, r2::Rect) -> dx, dy

Minimum distance or gap between two rectangles.

dx: The distance in the x-direction dy: The distance in the y-direction

The minimum distance will be zero when the rectangles are overlapping in a direction.

Rectangle.node_print

Rectangle.notvoid

Rectangle.olines

Rectangle.overlaps

Rectangle.parallelogram_area — Function.

    parallelogram_area(m::Matrix) -> Number

Area of the parallelogram. The matrix is a 2x3 matrix.

Rectangle.pcTol

Rectangle.perimeter

Rectangle.project

Rectangle.projectX — Function.

    projectX(r1::Rect, r2::Rect) -> (left, overlap, right)

Projects the rectangles along the X-axis and returns three parts of rectangles.

left: The left segment of the projection overlap: If there is any overlap between the rectangles right: The right segment of the projection

Each portion is returned as a tuple.

If the rectangle is part of the first rectangle, it's returned as the first element of the tuple. nothing is returned for a part when a portion is not available.

Rectangle.projectY — Function.

    projectY(r1::Rect, r2::Rect) -> (bottom, overlap, top)

Projects the rectangles along the Y-axis and returns three parts of rectangles.

bottom: The bottom segment of the projection overlap: If there is any overlap between the rectangles top: The top segment of the projection

Each portion is returned as a tuple.

If the rectangle is part of the first rectangle, it's returned as the first element of the tuple. nothing is returned for a part when a portion is not available.

Rectangle.ratio — Function.

    ratio(l1::Line{T}, p::Vector{T}) -> r::Real

If p is on l1 it divides the line at ratio r:(1-r) else nothing.

Rectangle.right_rotate!

Rectangle.ru

Rectangle.rx

Rectangle.ry

Rectangle.sortr

Rectangle.start

Rectangle.sx

Rectangle.sy

Rectangle.to_plot_shape

Rectangle.union

Rectangle.update_submax!

Rectangle.vert_asc — Function.

`isless` function  that can be used to sort vertical lines in ascending
order (left to right).

Rectangle.visible

Rectangle.visibleX — Function.

    visibleX(r1::Rect, r2::Rect) -> Rect
    visibleY(r1::Rect, r2::Rect) -> Rect

Projects the rectangles along the X-axis (Y-axis) and returns a rectangle area which is completely visible from both rectangles.

nothing is returned when there is no overlap along the X-axis.

Rectangle.visibleY

Rectangle.vline_xsection — Function.

Given a Rectangle and a set of already sorted set of vertical lines ordered
left to right, provides the indices that intersect the rectangle.

Rectangle.vlines

Rectangle.w

Rectangle.x

Rectangle.xsort

Rectangle.y

Rectangle.ysort