Docstrings

ReadOnlyArrays.ReadOnlyArray — Type.

ReadOnlyArray(a)

Return a read-only view into the parent array a. Most iteration, indexing, abstract arrays and strided arrays interface functions defined for array a are transparently defined for ReadOnlyArray(a). The exceptions are setindex! which is not allowed for ReadOnlyArray and similar which uses the default definitions for AbstractArray from Base. Also when used in broadcasting ReadOnlyArray uses default broadcast machinery.

Use parent function to access the parent array of ReadOnlyArray.

Examples

julia> a = [1 2; 3 4]
2×2 Array{Int64,2}:
 1  2
 3  4

julia> r = ReadOnlyArray(a)
2×2 ReadOnlyArray{Int64,2,Array{Int64,2}}:
 1  2
 3  4

julia> r[1]
1

julia> r[1] = 10
ERROR: setindex! not defined for ReadOnlyArray{Int64,2,Array{Int64,2}}
[...]

ReadOnlyArrays.ReadOnlyArrays

ReadOnlyArrays.eval

ReadOnlyArrays.include