Docstrings

BeliefUpdaters.BeliefUpdaters

BeliefUpdaters.DiscreteBelief — Type.

DiscreteBelief

A belief specified by a probability vector.

Normalization of b is NOT enforced at all times, but the DiscreteBeleif(pomdp, b) constructor will warn, and update(...) always returns a belief with normalized b.

BeliefUpdaters.DiscreteUpdater

BeliefUpdaters.FastPreviousObservationUpdater

BeliefUpdaters.KMarkovUpdater — Type.

Updater that stores the k most recent observations as the belief.

BeliefUpdaters.NothingUpdater

BeliefUpdaters.PreviousObservationUpdater — Type.

Updater that stores the most recent observation as the belief, missing if none is available.

BeliefUpdaters.PrimedPreviousObservationUpdater

BeliefUpdaters.create_belief

BeliefUpdaters.eval

BeliefUpdaters.include

BeliefUpdaters.throw_example

BeliefUpdaters.uniform_belief — Function.

Return a DiscreteBelief with equal probability for each state.