Docstrings

FixedEffects.AbstractHistogram

FixedEffects.AbstractWeights

FixedEffects.Adivtype

FixedEffects.AnalyticWeights

FixedEffects.CholeskyFixedEffectMatrix

FixedEffects.CoefTable

FixedEffects.ConvergenceException

FixedEffects.ConvergenceHistory

FixedEffects.ECDF

FixedEffects.FixedEffect

FixedEffects.FixedEffectMatrix

FixedEffects.FixedEffectVector

FixedEffects.FixedEffects

FixedEffects.FrequencyWeights

FixedEffects.Histogram

FixedEffects.L1dist

FixedEffects.L2dist

FixedEffects.LSMRFixedEffectMatrix

FixedEffects.LSMRParallelFixedEffectMatrix

FixedEffects.LSMRThreadslFixedEffectMatrix

FixedEffects.Linfdist

FixedEffects.ProbabilityWeights

FixedEffects.QRFixedEffectMatrix

FixedEffects.RegressionModel

FixedEffects.StatisticalModel

FixedEffects.StatsBase

FixedEffects.Weights

FixedEffects._LSMRFixedEffectMatrix

FixedEffects._solve_coefficients!

FixedEffects.addcounts!

FixedEffects.adjr2

FixedEffects.adjr²

FixedEffects.aic

FixedEffects.aicc

FixedEffects.autocor

FixedEffects.autocor!

FixedEffects.autocov

FixedEffects.autocov!

FixedEffects.aweights

FixedEffects.bic

FixedEffects.cache

FixedEffects.coef

FixedEffects.coefnames

FixedEffects.coeftable

FixedEffects.competerank

FixedEffects.confint

FixedEffects.connectedcomponent

FixedEffects.connectedcomponent!

FixedEffects.cor2cov

FixedEffects.corkendall

FixedEffects.corspearman

FixedEffects.counteq

FixedEffects.countmap

FixedEffects.countne

FixedEffects.counts

FixedEffects.cov2cor

FixedEffects.crosscor

FixedEffects.crosscor!

FixedEffects.crosscov

FixedEffects.crosscov!

FixedEffects.crossentropy

FixedEffects.denserank

FixedEffects.describe

FixedEffects.deviance

FixedEffects.dof

FixedEffects.dof_residual

FixedEffects.ecdf

FixedEffects.entropy

FixedEffects.eval

FixedEffects.factorize!

FixedEffects.findat

FixedEffects.fit

FixedEffects.fit!

FixedEffects.fitted

FixedEffects.fweights

FixedEffects.genmean

FixedEffects.genvar

FixedEffects.geomean

FixedEffects.get_fes

FixedEffects.gkldiv

FixedEffects.group

FixedEffects.harmmean

FixedEffects.helperC!

FixedEffects.helperN!

FixedEffects.hist

FixedEffects.include

FixedEffects.indexmap

FixedEffects.indicatormat

FixedEffects.informationmatrix

FixedEffects.initialize_refs

FixedEffects.initialize_where

FixedEffects.inverse_rle

FixedEffects.iqr

FixedEffects.isfitted

FixedEffects.islinear

FixedEffects.kldivergence

FixedEffects.kurtosis

FixedEffects.levelsmap

FixedEffects.leverage

FixedEffects.loglikelihood

FixedEffects.lsmr!

FixedEffects.mad

FixedEffects.mad!

FixedEffects.maxad

FixedEffects.mean_and_cov

FixedEffects.mean_and_std

FixedEffects.mean_and_var

FixedEffects.meanad

FixedEffects.meanresponse

FixedEffects.middle

FixedEffects.midpoints

FixedEffects.mode

FixedEffects.model_response

FixedEffects.modelmatrix

FixedEffects.modes

FixedEffects.moment

FixedEffects.msd

FixedEffects.mss

FixedEffects.nobs

FixedEffects.norepeats

FixedEffects.nquantile

FixedEffects.nulldeviance

FixedEffects.nullloglikelihood

FixedEffects.ordinalrank

FixedEffects.pacf

FixedEffects.pacf!

FixedEffects.partialcor

FixedEffects.percentile

FixedEffects.pool_combine!

FixedEffects.predict

FixedEffects.predict!

FixedEffects.proportionmap

FixedEffects.proportions

FixedEffects.psnr

FixedEffects.pweights

FixedEffects.r2

FixedEffects.renyientropy

FixedEffects.rescale!

FixedEffects.residuals

FixedEffects.response

FixedEffects.rle

FixedEffects.rmsd

FixedEffects.rss

FixedEffects.r²

FixedEffects.safe_rmul!

FixedEffects.sample

FixedEffects.sample!

FixedEffects.samplepair

FixedEffects.scattermat

FixedEffects.score

FixedEffects.sem

FixedEffects.skewness

FixedEffects.solve!

FixedEffects.solve_coefficients! — Function.

Solve a least square problem for a set of FixedEffects

solve_coefficients!(y, fes, weights; method = :lsmr, maxiter = 10000, tol = 1e-8)

Arguments

y : A AbstractVector{Float64}

fes: A Vector{<:FixedEffect}

weights: A AbstractWeights

method : A Symbol for the method. Choices are :lsmr, :lsmrthreads, :lsmrparallel, :qr and :cholesky

maxiter : Maximum number of iterations

tol : Tolerance

Returns

b : Solution of the least square problem

iterations: Number of iterations

converged: Did the algorithm converge?

Examples

using  FixedEffects
p1 = repeat(1:5, inner = 2)
p2 = repeat(1:5, outer = 2)
x = rand(10)
solve_coefficients!(rand(10), [FixedEffect(p1), FixedEffect(p2)])

FixedEffects.solve_residuals! — Function.

Solve a least square problem for a set of FixedEffects

solve_residuals!(y, fes, weights; method = :lsmr, maxiter = 10000, tol = 1e-8)

Arguments

y : A AbstractVector{Float64} or a AbstractMatrix{Float64}

fes: A Vector{<:FixedEffect}

weights: A AbstractWeights

method : A Symbol for the method. Choices are :lsmr, :lsmrthreads, :lsmrparallel, :qr and :cholesky

maxiter : Maximum number of iterations

tol : Tolerance

Returns

res : Residual of the least square problem

iterations: Number of iterations

converged: Did the algorithm converge?

Examples

using  FixedEffects
p1 = repeat(1:5, inner = 2)
p2 = repeat(1:5, outer = 2)
solve_residuals!(rand(10), [FixedEffect(p1), FixedEffect(p2)])
solve_residuals!(rand(10, 5), [FixedEffect(p1), FixedEffect(p2)])

FixedEffects.span

FixedEffects.sqL2dist

FixedEffects.stderror

FixedEffects.summarystats

FixedEffects.tiedrank

FixedEffects.totalvar

FixedEffects.trim

FixedEffects.trim!

FixedEffects.trimvar

FixedEffects.variation

FixedEffects.vcov

FixedEffects.weights

FixedEffects.winsor

FixedEffects.winsor!

FixedEffects.wmean

FixedEffects.wmean!

FixedEffects.wmedian

FixedEffects.wquantile

FixedEffects.wsample

FixedEffects.wsample!

FixedEffects.wsum

FixedEffects.wsum!

FixedEffects.zerox

FixedEffects.zscore

FixedEffects.zscore!