Docstrings

BytePairEncoding.AbstractNormalizer

BytePairEncoding.BPELearner

BytePairEncoding.Bpe

BytePairEncoding.BytePairEncoding

BytePairEncoding.Statistic

BytePairEncoding.UnNormalizer

BytePairEncoding.UtfNormalizer

BytePairEncoding._add_se_sym — Function.

add seperator and end symbol

BytePairEncoding._compute_options

BytePairEncoding._regex_escape — Function.

regex escape from https://github.com/JuliaLang/julia/pull/29643

BytePairEncoding.add! — Function.

add a new file to learner

BytePairEncoding.adjacent_pairs — Function.

get the adjacent pair of merged pair

BytePairEncoding.bi_pairs — Function.

find adjacent characters. return a list of Pair

find adjacent element. return a list of Pair

if only length one, return []

BytePairEncoding.emit — Function.

emit the BPE map to ofile; can add one-line comment to the header(first line)

BytePairEncoding.eval

BytePairEncoding.get_freq — Function.

get pair freq

get word freq

BytePairEncoding.get_vocab — Function.

get vocab with frequency counts

BytePairEncoding.include

BytePairEncoding.isolate_gloss — Function.

isolate given glossary in a word

BytePairEncoding.learn! — Function.

learn a BPE map

BytePairEncoding.lowestpair — Function.

find pair with lowest rank

BytePairEncoding.merge_pair! — Function.

merge the gived pair and update Statistic

BytePairEncoding.merged_pairs — Function.

merge the pair in the word tuple

BytePairEncoding.most_freq — Function.

most frequently pair

BytePairEncoding.normalize

BytePairEncoding.process_line — Function.

process a line, remain leading & trailing whitespace

BytePairEncoding.segment — Function.

segment a given sentence

BytePairEncoding.segment_token — Function.

bpe tokens

bpe a token and add seperator

BytePairEncoding.set_endsym — Function.

set the default end symbol

BytePairEncoding.set_tokenizer

BytePairEncoding.toStrTuple — Function.

toStrTuple(x; ends)

Turn string to a Tuple of each character in string. The last char will be concatenate with a ends. fefault ends = "</w>"

#Example

julia> x = toStrTuple("abcdef"; ends = "</w>")
("a", "b", "c", "d", "e", "f</w>")

BytePairEncoding.tokenize

BytePairEncoding.update! — Function.

update Statistic of the new merged pair

update Statistic of the removed adjacent pairs

update a tuple

update a word

update a frequency vocab

BytePairEncoding.whitespace_tokenize — Function.

simply the built-in split function for the origin tokenize method in subword-nmt