Docstrings

RadiationSpectra.AbstractFitFunction — Type.

AbstractFitFunction{T, N}

Abstract type for an N-dimensional fit of eltype T.

RadiationSpectra.FitFunction — Type.

FitFunction{T} <: AbstractFitFunction{T, 1}

Fields:

model::Function: Function of the fit model.

fitrange::Tuple{Union{Missing, T}, Union{Missing, T}}: Range on which the fit is performed.

parameters::Union{Vector{Missing}, Vector{T}}: Fitted parameters.

uncertainties::Union{Vector{Missing}, Vector{T}}: Uncertainties of the fitted parameters.

initial_parameters::Union{Vector{Missing}, Vector{T}}: Initial parameters.

confidence_level::Union{Missing, T}: Confidence level, used to determine the uncertainties.

Plotting:

A plot recipes exists for this struct: plot(fit::FitFunction{T}; npoints=501)

Plots the model function over the fitrange with 501-points.

RadiationSpectra.RadiationSpectra

RadiationSpectra.calculate_ratios

RadiationSpectra.calibrate_spectrum — Function.

calibrate_spectrum(h_uncal::Histogram, photon_lines::Array{Real, 1}; <keyword arguments>)::Histogram

Returns a calibrated histogram.

Keywords

sigma::Real = 2.0: The expected sigma of a peak in the spectrum. In units of bins.

threshold::Real = 10.0: Threshold for being identified as a peak in the deconvoluted spectrum. A single bin is identified as a peak when its weight exceeds the threshold and the previous bin was not identified as an peak.

min_n_peaks::Int = 0: If the number of found peaks is smaller than min_n_peaks the functions lowers the parameter threshold until enough peaks are found.

α::Real = 0.005: = 0.5%. Acceptance level in the comparison of the peak position ratios in the peak indentification step. When the difference between the ratio of two found peak positions and the ratio of two photon lines (photon_lines) is smaller than α, the found peaks are identified as the two photon lines.

Calibrate the spectrum h_uncal. This is done by:

finding peaks through devoncolution

identifying them through comparison of the ratios of their positions with the ratios of the known lines

fitting all identified peaks (with a gaussian plus first order polynomial) to get their position more precisely

performe a linear fit (offset forced to 0) of these positions vs the true positions (lines) to get the calibration constant

RadiationSpectra.determine_calibration_constant_through_peak_fitting

RadiationSpectra.determine_calibration_constant_through_peak_ratios

RadiationSpectra.eval

RadiationSpectra.gauss — Function.

gauss(x::T, p::Vector{T})::T

A Gaussian with 3 parameters:

p[1]: Scale/Amplitude

p[2]: σ

p[3]: μ

gauss(x::Vector{T}, p::Vector{T})::Vector{T}

Maps x to gauss(x::T, p::Vector{T})::T

RadiationSpectra.gauss_plus_first_order_polynom

RadiationSpectra.get_example_spectrum — Function.

get_example_spectrum()::Histogram

Returns an uncalibrated radiation spectrum for testing and demonstrating purpose.

RadiationSpectra.include

RadiationSpectra.linear_function_fixed_offset_at_zero

RadiationSpectra.llhfit

RadiationSpectra.llhfit! — Function.

 llhfit!(fit::FitFunction, h::Histogram)::Nothing

Performs a log-likelihood fit of the model function fit.model and the initial parameters fit.initial_parameters on the histogram h in the range fit.fitrange. The determined parameters are stored in fit.parameters.

The likelihood for each individual bin is the Poission distribution.

There are no uncertainty estimations for this fit yet.

RadiationSpectra.lsqfit! — Function.

lsqfit!(fit::FitFunction, h::Histogram; estimate_uncertainties=false)::Nothing

Performs a Least Square Fit with the model fit.model and the initial parameters fit.initial_parameters on the histogram h in the range fit.fitrange. The determined parameters are stored in fit.parameters and the corresponding uncertainties in fit.uncertainties for the given confidence level fit.confidence_level.

The uncertainties are marginalizations of the covariance matrix determined by the LSQFit.jl package. They are only calculated if the keywort estimate_uncertainties is set to true.

See LsqFit.jl for more detail.

RadiationSpectra.peakfinder — Function.

peakfinder(h::Histogram; <keyword arguments>)::Tuple{Histogram, Array{Float64, 1}}

Returns a deconvoluted spectrum and an array of peak positions.

Keywords

sigma::Real=2.0: The expected sigma of a peak in the spectrum. In units of bins.

threshold::Real=10.0: Threshold for being identified as a peak in the deconvoluted spectrum. A single bin is identified as a peak when its weight exceeds the threshold and the previous bin was not identified as an peak.

backgroundRemove::Bool=true

deconIterations::Int=3

markov::Bool=true

averWindow::Int=3

Source

This function is basically a copy of TSpectrum::SearchHighRes from ROOT.

M.A. Mariscotti: A method for identification of peaks in the presence of background and its application to spectrum analysis. NIM 50 (1967), 309-320.

M. Morhac;, J. Kliman, V. Matouoek, M. Veselsky, I. Turzo.:Identification of peaks in multidimensional coincidence gamma-ray spectra. NIM, A443 (2000) 108-125.

Z.K. Silagadze, A new algorithm for automatic photopeak searches. NIM A 376 (1996), 451.