Docstrings

RayTraceEllipsoid.Ellipsoid — Type.

Ellipsoid(c::Vec, r::Vec, dir::Vec, open::Float64)

An ellipsoid with a center, `c`, and radii, `r`, as well as a direction (gets automatically normalized), `dir`, and an opening angle, `open`, creating a dome (or window), that the ellipsoid is defined in. Note that `open` is the angle between the dome's edge and the direction of the dome (so actually half the opening angle) and is defined in **some angular units** (using UnitfulAngles, for example: u"°").

Ellipsoid has 6 additional fields all relating to various spatial transformations that convert the ellipsoid to a unit-sphere and back again. These are all CoordinateTransformations.

Examples

For an ellipsoid upwards-pointing hemisphere with a center at (1,2,3), and radii (4,5,6):

julia> Ellipsoid(Vec(1,2,3), Vec(4,5,6), Vec(0,0,1), 0.0)
Rayden.Ellipsoid([1.0, 2.0, 3.0], [4.0, 5.0, 6.0], [0.0, 0.0, 1.0], 0.0, Translation(-1.0, -2.0, -3.0), LinearMap([0.25 0.0 0.0; 0.0 0.2 0.0; 0.0 0.0 0.166667]), AffineMap([0.25 0.0 0.0; 0.0 0.2 0.0; 0.0 0.0 0.166667], [-0.25, -0.4, -0.5]), Translation(1.0, 2.0, 3.0), LinearMap([4.0 0.0 0.0; 0.0 5.0 0.0; 0.0 0.0 6.0]), AffineMap([4.0 0.0 0.0; 0.0 5.0 0.0; 0.0 0.0 6.0], [1.0, 2.0, 3.0]))

RayTraceEllipsoid.Interface — Type.

Interface(normal::AffineMap, n::Float64)

Build an optical interface from a AffineMap that transforms a point on the ellipsoid of the interface to the normal at that point, normal, and the refractive index ratio between the inside and the outside of the ellipsoid, n.

RayTraceEllipsoid.OpticUnit — Type.

OpticUnit(body::Ellipsoid, interface::Interface, register::Bool, name::String)

Build an optical unit. register indicates whether this unit should register intersection points (thus functions as a retina).

RayTraceEllipsoid.Ray — Type.

Ray(orig::Vec, dir::Vec)

The main Ray type with a ray origin, orig, and direction, dir. The ray's direction gets normalized.

RayTraceEllipsoid.RayTraceEllipsoid

RayTraceEllipsoid.Vec — Type.

Vec = SVector{3,Float64}

Points and directions are just a point in 3D space best described as a static vector, SVector. Vec is an alias for that.

RayTraceEllipsoid.advance! — Function.

advance!(r::Ray, s::Ellipsoid)

Find the shortest point of intersection that is within the ellipsoid's dome and reassign the origin of the ray. Returns failure of the intersection.

RayTraceEllipsoid.bend! — Function.

bend!(r::Ray, i::Interface)

Refract or reflect a ray with an interface. Update the direction of the ray. Returns if event failed, which is always false (see raytrace! for details why that is so).

RayTraceEllipsoid.distance — Function.

distance(orig::Vec, dir::Vec)

Return the two distances between a point with origin, orig, and direction, dir, and the two (potentially identical) intersection points with a unit-sphere.

RayTraceEllipsoid.eval

RayTraceEllipsoid.include

RayTraceEllipsoid.raytrace! — Function.

raytrace!(r::Ray, c::OpticUnit)

Advance a ray to the intersection point with an ellipsoid. If the intersection was successful, bend the ray. Updates the ray accordingly. Returns intersection failure.