Docstrings

Mixers.@mix — Macro.

Generates a mixin macro that preserve parametric types. Fields and parametric types are appended to the struct. Identical parametric types are merged.

Mixers.@pour — Macro.

Generates simple macros for inserting any lines anywhere.

Mixers.@premix — Macro.

Just like @mix but generated macro insert fields and types at the start of the definition.

Mixers.Mixers

Mixers.chained_macros

Mixers.chained_macros!

Mixers.defmacro

Mixers.eval

Mixers.firsthead

Mixers.include

Mixers.mergefields

Mixers.mergetypes

Mixers.mix