Docstrings

BasisFunctionExpansions.AB_covariance

BasisFunctionExpansions.BasisFunctionApproximation — Type.

BasisFunctionApproximation(y::Vector, v, bfe::BasisFunctionExpansion, λ = 0)

Perform parameter identification to identify the Function y = ϕ(v), where ϕ is a Basis Function Expansion of type bfe. λ is an optional regularization parameter (L² regularization).

BasisFunctionExpansions.BasisFunctionExpansion

BasisFunctionExpansions.BasisFunctionExpansions

BasisFunctionExpansions.LPVSS — Type.

Convenience tyoe for estimation of LPV state-space models

BasisFunctionExpansions.MultiDiagonalRBFE — Type.

A MultiDiagonalRBFE has different diagonal covariance matrices for all basis functions See also MultiUniformRBFE, which has the same covariance matrix for all basis functions

BasisFunctionExpansions.MultiRBFE — Type.

A MultiRBFE has different diagonal covariance matrices for all basis functions See also MultiUniformRBFE, which has the same covariance matrix for all basis functions

BasisFunctionExpansions.MultiUniformRBFE — Type.

A MultiUniformRBFE has the same diagonal covariance matrix for all basis functions See also MultiDiagonalRBFE, which has different covariance matrices for all basis functions

BasisFunctionExpansions.UniformRBFE — Type.

A Uniform RBFE has the same variance for all basis functions

BasisFunctionExpansions.basis_activation_func

BasisFunctionExpansions.basis_activation_func_automatic — Function.

basis_activation_func_automatic(v,Nv,normalize,coulomb)

Returns a func v->ϕ(v) ∈ ℜ(Nv) that calculates the activation of Nv basis functions spread out to cover v nicely. If coulomb is true, then we get twice the number of basis functions, 2Nv, with a hard split at v=0 (useful to model Coulomb friction). coulomb is not yet fully supported for all expansion types.

The keyword normalize determines weather or not basis function activations are normalized to sum to one for each datapoint, normalized networks tend to extrapolate better "The normalized radial basis function neural network" DOI: 10.1109/ICSMC.1998.728118

BasisFunctionExpansions.eval

BasisFunctionExpansions.fit_ss

BasisFunctionExpansions.getARXregressor — Function.

getARXregressor(y::AbstractVector,u::AbstractVecOrMat, na, nb)

Returns a shortened output signal y and a regressor matrix A such that the least-squares ARX model estimate of order na,nb is y\A

Return a regressor matrix used to fit an ARX model on, e.g., the form A(z)y = B(z)f(u) with output y and input u where the order of autoregression is na and the order of input moving average is nb

Example

Here we test the model with the Function f(u) = √(|u|)

A     = [1,2*0.7*1,1] # A(z) coeffs
B     = [10,5] # B(z) coeffs
u     = randn(100) # Simulate 100 time steps with Gaussian input
y     = filt(B,A,sqrt.(abs.(u)))
yr,A  = getARXregressor(y,u,3,2) # We assume that we know the system order 3,2
bfe   = MultiUniformRBFE(A,[2,2,4,4,4], normalize=true)
bfa   = BasisFunctionApproximation(yr,A,bfe, 1e-3)
e_bfe = √(mean((yr - bfa(A)).^2)) # (0.005174261451622258)
plot([yr bfa(A)], lab=["Signal" "Prediction"])

See README (?BasisFunctionExpansions) for more details

BasisFunctionExpansions.getARregressor — Function.

y,A = getARregressor(y::AbstractVector,na::Integer)

Returns a shortened output signal y and a regressor matrix A such that the least-squares AR model estimate of order na is y\A

BasisFunctionExpansions.get_centers — Function.

vc,γ = get_centers(bounds, Nv, coulomb=false, coulombdims=0)

BasisFunctionExpansions.get_centers_Kmeans

BasisFunctionExpansions.get_centers_automatic — Function.

vc,γ = get_centers_automatic(v::AbstractVector,Nv::Int,coulomb = false)

vc,γ = get_centers_automatic(v::AbstractMatrix, Nv::AbstractVector{Int}, coulomb=false, coulombdims=0)

BasisFunctionExpansions.get_centers_multi

BasisFunctionExpansions.include

BasisFunctionExpansions.matricesn — Function.

size(y) = (T-1, n)

BasisFunctionExpansions.mega_regressor

BasisFunctionExpansions.meshgrid

BasisFunctionExpansions.normalized_squared_exponential

BasisFunctionExpansions.normalized_squared_exponential_coulomb

BasisFunctionExpansions.output_variance — Function.

output_variance(model::LPVSS, x::AbstractVector, u::AbstractVector, v=[x u])

Return a vector of prediction variances. Note, no covariance between dimensions in output is provided

BasisFunctionExpansions.predict — Function.

predict(model::LPVSS, x::AbstractMatrix, u, v=[x u])

If no v provided, return a prediction of the output x' given the state x and input u This function is called when a model::LPVSS object is called like model(x,u)

Provided v, return a prediction of the output x' given the state x, input u and scheduling parameter v

BasisFunctionExpansions.quadform

BasisFunctionExpansions.shorten_v

BasisFunctionExpansions.squared_exponential

BasisFunctionExpansions.squared_exponential_coulomb

BasisFunctionExpansions.testdata — Function.

x,xm,u,n,m = testdata(T,r=1)

Generate T time steps of state-space data where the A-matrix changes from A = [0.95 0.1; 0 0.95] to A = [0.5 0.05; 0 0.5] at time t=T÷2 x,xm,u,n,m = (state,noisy state, input, statesize, inputsize) r is the seed to the random number generator.

BasisFunctionExpansions.toeplitz — Function.

toeplitz{T}(c::AbstractArray{T},r::AbstractArray{T})

Returns a Toeplitz matrix where c is the first column and r is the first row.

BasisFunctionExpansions.γ2σ

BasisFunctionExpansions.σ2γ