Docstrings

TerminalMenus.ARROW_DOWN

TerminalMenus.ARROW_LEFT

TerminalMenus.ARROW_RIGHT

TerminalMenus.ARROW_UP

TerminalMenus.AbstractMenu — Type.

AbstractMenu

The supertype for all Menu types. See AbstractMenu.jl for descriptions of functions mentioned in this doc string.

Functions

The following functions can be called on all <:AbstractMenu types.

Exported

request(m::AbstractMenu)

request(msg::AbstractString, m::AbstractMenu)

Hidden

printMenu(m::AbstractMenu, cursor::Int; init::Bool=false)

Subtypes

All subtypes must contain the feilds pagesize::Int and pageoffset::Int. They must also implement the following functions.

Necessary Functions

These functions must be implemented for all subtypes of AbstractMenu.

pick(m::AbstractMenu, cursor::Int)

cancel(m::AbstractMenu)

options(m::AbstractMenu)

writeLine(buf::IOBuffer, m::AbstractMenu, idx::Int, cur::Bool, term_width::Int)

Optional Functions

These functions do not need to be implemented for all AbstractMenu subtypes.

header(m::AbstractMenu)

keypress(m::AbstractMenu, i::UInt32)

TerminalMenus.CONFIG — Constant.

global menu configuration parameters

TerminalMenus.DEL_KEY

TerminalMenus.END_KEY

TerminalMenus.HOME_KEY

TerminalMenus.Key

TerminalMenus.MultiSelectMenu — Type.

MultiSelectMenu

A menu that allows a user to select a multiple options from a list.

Sample Output

julia> request(MultiSelectMenu(options))
Select the fruits you like:
[press: d=done, a=all, n=none, <enter>=select]
   [ ] apple
 > [X] orange
   [X] grape
   [ ] strawberry
   [ ] blueberry
   [X] peach
   [ ] lemon
   [ ] lime
You like the following fruits:
  - orange
  - grape
  - peach

TerminalMenus.PAGE_DOWN

TerminalMenus.PAGE_UP

TerminalMenus.RadioMenu — Type.

RadioMenu

A menu that allows a user to select a single option from a list.

Sample Output

julia> request(RadioMenu(options, pagesize=4))
Choose your favorite fruit:
^  grape
   strawberry
 > blueberry
v  peach
Your favorite fruit is blueberry!

TerminalMenus.TerminalMenus

TerminalMenus.__init__

TerminalMenus.cancel

TerminalMenus.config — Function.

config( <see arguments> )

Keyword-only function to configure global menu parameters

Arguments

charset::Symbol=:na: ui characters to use (:ascii or :unicode); overridden by other arguments

cursor::Char='>'|'→': character to use for cursor

up_arrow::Char='^'|'↑': character to use for up arrow

down_arrow::Char='v'|'↓': character to use for down arrow

checked::String="[X]"|"✓": string to use for checked

unchecked::String="[ ]"|"⬚"): string to use for unchecked

scroll::Symbol=:nowrap: If :wrap then wrap the cursor around top and bottom, if :nowrap do not wrap cursor

supress_output::Bool=false: For testing. If true, menu will not be printed to console.

ctrl_c_interrupt::Bool=true: If false, return empty on ^C, if true throw InterruptException() on ^C

TerminalMenus.disableRawMode

TerminalMenus.enableRawMode

TerminalMenus.eval

TerminalMenus.header

TerminalMenus.include

TerminalMenus.keypress

TerminalMenus.options

TerminalMenus.pick

TerminalMenus.printMenu

TerminalMenus.readKey

TerminalMenus.readNextChar

TerminalMenus.request — Function.

request(m::AbstractMenu)

Display the menu and enter interactive mode. Returns m.selected which varies based on menu type.

request([term,] msg::AbstractString, m::AbstractMenu)

Shorthand for println(msg); request(m).

TerminalMenus.terminal

TerminalMenus.trimWidth

TerminalMenus.writeLine