Docstrings

Diana.Argument

Diana.BooleanValue

Diana.Client

Diana.Diana

Diana.Directive

Diana.DirectiveDefinition

Diana.Document

Diana.EnumTypeDefinition

Diana.EnumValue

Diana.EnumValueDefinition

Diana.ExecuteQuery

Diana.Field

Diana.FieldDefinition

Diana.FloatValue

Diana.FragmentCycles

Diana.FragmentDefinition

Diana.FragmentNames

Diana.FragmentSpread

Diana.FragmentSubscription

Diana.FragmentUnknowNotUsed

Diana.GraphQLClient

Diana.GraphqlError

Diana.InlineFragment

Diana.InputObjectTypeDefinition

Diana.InputValueDefinition

Diana.IntValue

Diana.InterfaceTypeDefinition

Diana.Lexer

Diana.Lexers

Diana.ListType

Diana.List_

Diana.Name

Diana.NamedType

Diana.Node

Diana.NonNullType

Diana.NotExtensionOnOperation

Diana.NotSchemaOnOperation

Diana.NotTypeOnOperation

Diana.NullValue

Diana.ObjectTypeDefinition

Diana.Object_

Diana.Object_Field

Diana.OperationAnonymous

Diana.OperationDefinition

Diana.OperationNames

Diana.OperationTypeDefinition

Diana.Parse — Function.

Given a GraphQL source, parses it into a Document. Throws GraphQLError if a syntax error is encountered.

Diana.Queryclient

Diana.Result

Diana.ScalarTypeDefinition

Diana.Schema

Diana.SchemaDefinition

Diana.SelectionSet

Diana.StringValue

Diana.SubscriptionFields

Diana.Lexers.Tokenize — Function.

Tokenize(x)

Returns an Iterable containing the tokenized input. Can be reverted by e.g. join(untokenize.(Tokenize(x))).

Diana.Tokens

Diana.Tokensgraphql

Diana.TypeExtensionDefinition

Diana.UnionTypeDefinition

Diana.Validatequery

Diana.Variable

Diana.VariableDefinition

Diana.an_y — Function.

Returns a possibly empty list of parse nodes, determined by the parseFn. This list begins with a lex token of openKind and ends with a lex token of closeKind. Advances the parser to the next lex token after the closing token.

Diana.createLexer

Diana.deepquery

Diana.eval

Diana.executor

Diana.expect — Function.

If the next token is of the given kind, return that token after advancing the lexer. Otherwise, do not change the parser state and throw an error.

Diana.expectKeyword — Function.

If the next token is a keyword with the given value, return that token after advancing the lexer. Otherwise, do not change the parser state and return false.

Diana.getdeep

Diana.getfield_types

Diana.gettypes

Diana.include

Diana.many — Function.

Returns a non-empty list of parse nodes, determined by the parseFn. This list begins with a lex token of openKind and ends with a lex token of closeKind. Advances the parser to the next lex token after the closing token.

Diana.next_token — Function.

If the next token is of the given kind, return true after advancing the lexer. Otherwise, do not change the parser state and return false.

Diana.parseArgument — Function.

Argument : Name : Value

Diana.parseArgumentDefs — Function.

ArgumentsDefinition : ( InputValueDefinition+ )

Diana.parseArguments — Function.

Arguments : ( Argument+ )

Diana.parseConstValue

Diana.parseDefinition — Function.

Definition : - OperationDefinition - FragmentDefinition - TypeSystemDefinition

Diana.parseDirective — Function.

Directive : @ Name Arguments?

Diana.parseDirectiveDefinition — Function.

DirectiveDefinition : - directive @ Name ArgumentsDefinition? on DirectiveLocations

Diana.parseDirectiveLocations — Function.

DirectiveLocations : - |? Name - DirectiveLocations | Name

Diana.parseDirectives — Function.

Implements the parsing rules in the Directives section.

Directives : Directive+

Diana.parseDocument — Function.

#if length(errors)>0
#	println("Syntax Error GraphQL request (x:x) Unexpected character")
#	errors
	# {"errors":[{"locations":[{"column":5,"line":3}],"message":"Syntax Error GraphQL request (3:5) Unexpected character "".

2: neomatrix{

3:     nombre
   ^

4: linkedin "}]} Document : Definition+

Diana.parseEnumTypeDefinition — Function.

EnumTypeDefinition : enum Name Directives? { EnumValueDefinition+ }

Diana.parseEnumValueDefinition — Function.

EnumValueDefinition : EnumValue Directives?

EnumValue : Name

Diana.parseField — Function.

Field : Alias? Name Arguments? Directives? SelectionSet?

Alias : Name : NAME

Diana.parseFieldDefinition — Function.

FieldDefinition : Name ArgumentsDefinition? : Type Directives?

Diana.parseFragment — Function.

Implements the parsing rules in the Fragments section.

Corresponds to both FragmentSpread and InlineFragment in the spec.

FragmentSpread : ... FragmentName Directives?

InlineFragment : ... TypeCondition? Directives? SelectionSet

Diana.parseFragmentDefinition — Function.

FragmentDefinition : - fragment FragmentName on TypeCondition Directives? SelectionSet

TypeCondition : NamedType

Diana.parseFragmentName — Function.

FragmentName : Name but not on

Diana.parseImplementsInterfaces — Function.

ImplementsInterfaces : implements NamedType+

Diana.parseInputObjectTypeDefinition — Function.

InputObjectTypeDefinition : input Name Directives? { InputValueDefinition+ }

Diana.parseInputValueDef — Function.

InputValueDefinition : Name : Type DefaultValue? Directives?

Diana.parseInterfaceTypeDefinition — Function.

InterfaceTypeDefinition : interface Name Directives? { FieldDefinition+ }

Diana.parseList — Function.

ListValue[Const] : - [ ] - [ Value[?Const]+ ]

Diana.parseName — Function.

Converts a name lex token into a name parse node.

Diana.parseNamedType — Function.

NamedType : Name

Diana.parseObject — Function.

ObjectValue[Const] : - { } - { ObjectField[?Const]+ }

Diana.parseObjectField — Function.

ObjectField[Const] : Name : Value[?Const]

Diana.parseObjectTypeDefinition — Function.

ObjectTypeDefinition : - type Name ImplementsInterfaces? Directives? { FieldDefinition+ }

Diana.parseOperationDefinition — Function.

Implements the parsing rules in the Operations section.

OperationDefinition :

SelectionSet

OperationType Name? VariableDefinitions? Directives? SelectionSet

Diana.parseOperationType — Function.

Implements the parsing rules in the Document section.

OperationType : one of query mutation subscription

Diana.parseOperationTypeDefinition

Diana.parseScalarTypeDefinition — Function.

ScalarTypeDefinition : scalar Name Directives?

Diana.parseSchemaDefinition — Function.

SchemaDefinition : schema Directives? { OperationTypeDefinition+ }

OperationTypeDefinition : OperationType : NamedType

Diana.parseSelection — Function.

Selection : - Field - FragmentSpread - InlineFragment

Diana.parseSelectionSet — Function.

SelectionSet : { Selection+ }

Diana.parseType — Function.

Given a string containing a GraphQL Type (ex. [Int!]), parse the AST for that type. Throws GraphQLError if a syntax error is encountered.

This is useful within tools that operate upon GraphQL Types directly and in isolation of complete GraphQL documents.

Consider providing the results to the utility function: typeFromAST().

Diana.parseTypeExtensionDefinition — Function.

TypeExtensionDefinition : extend ObjectTypeDefinition

Diana.parseTypeReference — Function.

Implements the parsing rules in the Types section.

Type : - NamedType - ListType - NonNullType

Diana.parseTypeSystemDefinition — Function.

Implements the parsing rules in the Type Definition section.

TypeSystemDefinition : - SchemaDefinition - TypeDefinition - TypeExtensionDefinition - DirectiveDefinition

TypeDefinition : - ScalarTypeDefinition - ObjectTypeDefinition - InterfaceTypeDefinition - UnionTypeDefinition - EnumTypeDefinition - InputObjectTypeDefinition

Diana.parseUnionMembers — Function.

UnionMembers :
- `|`? NamedType
- UnionMembers | NamedType

Diana.parseUnionTypeDefinition — Function.

UnionTypeDefinition : union Name Directives? = UnionMembers

Diana.parseValue — Function.

Given a string containing a GraphQL value (ex. [42]), parse the AST for that value. Throws GraphQLError if a syntax error is encountered.

This is useful within tools that operate upon GraphQL Values directly and in isolation of complete GraphQL documents.

Consider providing the results to the utility function: valueFromAST().

Diana.parseValueLiteral — Function.

Implements the parsing rules in the Values section.

Value[Const] : - [~Const] Variable - IntValue - FloatValue - StringValue - BooleanValue - NullValue - EnumValue - ListValue[?Const] - ObjectValue[?Const]

BooleanValue : one of true false

NullValue : null

EnumValue : Name but not true, false or null

Diana.parseValueValue

Diana.parseVariable — Function.

Variable : $ Name

Diana.parseVariableDefinition — Function.

VariableDefinition : Variable : Type DefaultValue?

Diana.parseVariableDefinitions — Function.

VariableDefinitions : ( VariableDefinition+ )

Diana.peek — Function.

Determines if the next token is of a given kind

Diana.resol

Diana.schema

Diana.unexpected — Function.

Helper function for creating an error when an unexpected lexed token is encountered.

Diana.visitante