Docstrings

TuringModels.+

TuringModels.-

TuringModels.@VarName

TuringModels.@animate

TuringModels.@colorant_str

TuringModels.@csv2_str

TuringModels.@csv_str

TuringModels.@df

TuringModels.@gif

TuringModels.@layout

TuringModels.@model

TuringModels.@recipe

TuringModels.@series

TuringModels.@shorthands

TuringModels.@tsv_str

TuringModels.@userplot

TuringModels.@wsv_str

TuringModels.ABGR

TuringModels.ADADelta

TuringModels.ADAGrad

TuringModels.ADAM

TuringModels.ADAMW

TuringModels.ADIN99

TuringModels.ADIN99d

TuringModels.ADIN99o

TuringModels.AGray

TuringModels.AGray32

TuringModels.AHSI

TuringModels.AHSL

TuringModels.AHSV

TuringModels.ALCHab

TuringModels.ALCHuv

TuringModels.ALMS

TuringModels.ALab

TuringModels.ALuv

TuringModels.AMSGrad

TuringModels.ARGB

TuringModels.ARGB32

TuringModels.AXYZ

TuringModels.AYCbCr

TuringModels.AYIQ

TuringModels.AbstractBackend

TuringModels.AbstractCategoricalArray

TuringModels.AbstractCategoricalMatrix

TuringModels.AbstractCategoricalVector

TuringModels.AbstractDataFrame

TuringModels.AbstractGray

TuringModels.AbstractHistogram

TuringModels.AbstractLayout

TuringModels.AbstractMissingCategoricalArray

TuringModels.AbstractMissingCategoricalMatrix

TuringModels.AbstractMissingCategoricalVector

TuringModels.AbstractMixtureModel

TuringModels.AbstractMvNormal

TuringModels.AbstractPlot

TuringModels.AbstractRGB

TuringModels.AbstractSerializer

TuringModels.AbstractWeights

TuringModels.AdaMax

TuringModels.AlphaColor

TuringModels.AnalyticWeights

TuringModels.Animation

TuringModels.Arcsine

TuringModels.AxyY

TuringModels.BGR

TuringModels.BGRA

TuringModels.BatchNorm

TuringModels.Bernoulli

TuringModels.Beta

TuringModels.BetaBinomial

TuringModels.BetaPrime

TuringModels.BezierCurve

TuringModels.Binomial

TuringModels.BinomialLogit

TuringModels.Biweight

TuringModels.CIE1931JV_CMF

TuringModels.CIE1931J_CMF

TuringModels.CIE1931_CMF

TuringModels.CIE1964_CMF

TuringModels.CSMC

TuringModels.CSV

TuringModels.CTask

TuringModels.Categorical

TuringModels.CategoricalArray

TuringModels.CategoricalArrays

TuringModels.CategoricalMatrix

TuringModels.CategoricalPool

TuringModels.CategoricalString

TuringModels.CategoricalValue

TuringModels.CategoricalVector

TuringModels.Cauchy

TuringModels.Chain

TuringModels.Chains

TuringModels.Chi

TuringModels.Chisq

TuringModels.CoefTable

TuringModels.Color

TuringModels.Color3

TuringModels.ColorAlpha

TuringModels.ColorGradient

TuringModels.ColorTypes

TuringModels.Colorant

TuringModels.ColorantNormed

TuringModels.Colors

TuringModels.Continuous

TuringModels.ContinuousDistribution

TuringModels.ContinuousMatrixDistribution

TuringModels.ContinuousMultivariateDistribution

TuringModels.ContinuousUnivariateDistribution

TuringModels.Conv

TuringModels.ConvergenceException

TuringModels.Cosine

TuringModels.DE_2000

TuringModels.DE_94

TuringModels.DE_AB

TuringModels.DE_BFD

TuringModels.DE_CMC

TuringModels.DE_DIN99

TuringModels.DE_DIN99d

TuringModels.DE_DIN99o

TuringModels.DE_JPC79

TuringModels.DIN99

TuringModels.DIN99A

TuringModels.DIN99d

TuringModels.DIN99dA

TuringModels.DIN99o

TuringModels.DIN99oA

TuringModels.DataFrame

TuringModels.DataFrameRow

TuringModels.DataFrames

TuringModels.DelimitedFiles

TuringModels.Dense

TuringModels.DepthwiseConv

TuringModels.Descent

TuringModels.DiagNormal

TuringModels.DiagNormalCanon

TuringModels.Dirichlet

TuringModels.DirichletMultinomial

TuringModels.Discrete

TuringModels.DiscreteDistribution

TuringModels.DiscreteMatrixDistribution

TuringModels.DiscreteMultivariateDistribution

TuringModels.DiscreteUniform

TuringModels.DiscreteUnivariateDistribution

TuringModels.Distribution

TuringModels.Distributions

TuringModels.DoubleExponential

TuringModels.Dropout

TuringModels.DynamicNUTS

TuringModels.ECDF

TuringModels.EdgeworthMean

TuringModels.EdgeworthSum

TuringModels.EdgeworthZ

TuringModels.EmpiricalUnivariateDistribution

TuringModels.Epanechnikov

TuringModels.Erlang

TuringModels.Estimator

TuringModels.ExpDecay

TuringModels.Exponential

TuringModels.FDist

TuringModels.FisherNoncentralHypergeometric

TuringModels.Flat

TuringModels.FlatPos

TuringModels.Flux

TuringModels.Formatted

TuringModels.Fractional

TuringModels.Frechet

TuringModels.FrequencyWeights

TuringModels.FullNormal

TuringModels.FullNormalCanon

TuringModels.GR

TuringModels.GRU

TuringModels.Gamma

TuringModels.GeneralizedExtremeValue

TuringModels.GeneralizedPareto

TuringModels.Geometric

TuringModels.Gibbs

TuringModels.Gray

TuringModels.Gray24

TuringModels.GrayA

TuringModels.GroupApplied

TuringModels.GroupedDataFrame

TuringModels.Gumbel

TuringModels.HMC

TuringModels.HMCDA

TuringModels.HSB

TuringModels.HSI

TuringModels.HSIA

TuringModels.HSL

TuringModels.HSLA

TuringModels.HSV

TuringModels.HSVA

TuringModels.Histogram

TuringModels.Hypergeometric

TuringModels.IPMCMC

TuringModels.IS

TuringModels.InvDecay

TuringModels.InverseGamma

TuringModels.InverseGaussian

TuringModels.InverseWishart

TuringModels.IsoNormal

TuringModels.IsoNormalCanon

TuringModels.KSDist

TuringModels.KSOneSided

TuringModels.KW

TuringModels.Kolmogorov

TuringModels.L1dist

TuringModels.L2dist

TuringModels.LCHab

TuringModels.LCHabA

TuringModels.LCHuv

TuringModels.LCHuvA

TuringModels.LMS

TuringModels.LMSA

TuringModels.LSTM

TuringModels.Lab

TuringModels.LabA

TuringModels.Laplace

TuringModels.LayerNorm

TuringModels.LevelsException

TuringModels.Levy

TuringModels.Libtask

TuringModels.Linfdist

TuringModels.LocationScale

TuringModels.LogNormal

TuringModels.Logistic

TuringModels.Luv

TuringModels.LuvA

TuringModels.MCMCChain

TuringModels.MH

TuringModels.MLEstimator

TuringModels.MSC

TuringModels.MatrixDistribution

TuringModels.Matrixvariate

TuringModels.MaxPool

TuringModels.MeanPool

TuringModels.Missing

TuringModels.MissingCategoricalArray

TuringModels.MissingCategoricalMatrix

TuringModels.MissingCategoricalVector

TuringModels.MissingException

TuringModels.Missings

TuringModels.MixtureModel

TuringModels.Momentum

TuringModels.Multinomial

TuringModels.Multivariate

TuringModels.MultivariateDistribution

TuringModels.MultivariateMixture

TuringModels.MultivariateNormal

TuringModels.MvLogNormal

TuringModels.MvNormal

TuringModels.MvNormalCanon

TuringModels.MvNormalKnownCov

TuringModels.MvTDist

TuringModels.NADAM

TuringModels.NNlib

TuringModels.NUTS

TuringModels.NegativeBinomial

TuringModels.Nesterov

TuringModels.NonMatrixDistribution

TuringModels.NoncentralBeta

TuringModels.NoncentralChisq

TuringModels.NoncentralF

TuringModels.NoncentralHypergeometric

TuringModels.NoncentralT

TuringModels.Normal

TuringModels.NormalCanon

TuringModels.NormalInverseGaussian

TuringModels.OHLC

TuringModels.OrderedLevelsException

TuringModels.PG

TuringModels.PIMH

TuringModels.PMMH

TuringModels.Pareto

TuringModels.PlotThemes

TuringModels.PlotUtils

TuringModels.Plots

TuringModels.Poisson

TuringModels.PoissonBinomial

TuringModels.ProbabilityWeights

TuringModels.QQPair

TuringModels.RDatasets

TuringModels.RGB

TuringModels.RGB1

TuringModels.RGB24

TuringModels.RGB4

TuringModels.RGBA

TuringModels.RMSProp

TuringModels.RNN

TuringModels.Rayleigh

TuringModels.RealInterval

TuringModels.RecipeData

TuringModels.RecipesBase

TuringModels.RegressionModel

TuringModels.SGD

TuringModels.SGHMC

TuringModels.SGLD

TuringModels.SMC

TuringModels.Sampleable

TuringModels.ScriptEntry

TuringModels.Segments

TuringModels.Semicircle

TuringModels.Serialization

TuringModels.Serializer

TuringModels.Shape

TuringModels.Skellam

TuringModels.StatisticalModel

TuringModels.StatisticalRethinking

TuringModels.StatsBase

TuringModels.StatsFuns

TuringModels.StatsPlots

TuringModels.SubDataFrame

TuringModels.SufficientStats

TuringModels.Surface

TuringModels.SymTriangularDist

TuringModels.TArray

TuringModels.TDist

TuringModels.TRef

TuringModels.TrackedArray

TuringModels.TrackedMatrix

TuringModels.TrackedVector

TuringModels.Tracker

TuringModels.Transparent3

TuringModels.TransparentColor

TuringModels.TransparentGray

TuringModels.TransparentRGB

TuringModels.TriangularDist

TuringModels.Triweight

TuringModels.Truncated

TuringModels.TruncatedNormal

TuringModels.Turing

TuringModels.TuringModels

TuringModels.Uniform

TuringModels.Univariate

TuringModels.UnivariateDistribution

TuringModels.UnivariateGMM

TuringModels.UnivariateMixture

TuringModels.ValueSupport

TuringModels.VariateForm

TuringModels.VecBinomialLogit

TuringModels.VonMises

TuringModels.VonMisesFisher

TuringModels.WalleniusNoncentralHypergeometric

TuringModels.Weibull

TuringModels.WeightDecay

TuringModels.Weights

TuringModels.Wishart

TuringModels.XYZ

TuringModels.XYZA

TuringModels.YCbCr

TuringModels.YCbCrA

TuringModels.YIQ

TuringModels.YIQA

TuringModels.ZeroMeanDiagNormal

TuringModels.ZeroMeanDiagNormalCanon

TuringModels.ZeroMeanFullNormal

TuringModels.ZeroMeanFullNormalCanon

TuringModels.ZeroMeanIsoNormal

TuringModels.ZeroMeanIsoNormalCanon

TuringModels.adapted_grid

TuringModels.add_backend

TuringModels.add_theme

TuringModels.addcounts!

TuringModels.adjr2

TuringModels.adjr²

TuringModels.aggregate

TuringModels.aic

TuringModels.aicc

TuringModels.aliases

TuringModels.allowmissing

TuringModels.allowmissing!

TuringModels.alpha

TuringModels.alphacolor

TuringModels.andrewsplot

TuringModels.andrewsplot!

TuringModels.animate

TuringModels.annotate!

TuringModels.arrow

TuringModels.attr!

TuringModels.autcorplot

TuringModels.auto_tune_chunk_size!

TuringModels.autocor

TuringModels.autocor!

TuringModels.autocorplot

TuringModels.autocorplot!

TuringModels.autocov

TuringModels.autocov!

TuringModels.aweights

TuringModels.backend

TuringModels.backend_name

TuringModels.backend_object

TuringModels.backends

TuringModels.bar

TuringModels.bar!

TuringModels.barh

TuringModels.barh!

TuringModels.barhist

TuringModels.barhist!

TuringModels.base_color_type

TuringModels.base_colorant_type

TuringModels.bbox

TuringModels.betaccdf

TuringModels.betacdf

TuringModels.betainvccdf

TuringModels.betainvcdf

TuringModels.betainvlogccdf

TuringModels.betainvlogcdf

TuringModels.betalogccdf

TuringModels.betalogcdf

TuringModels.betalogpdf

TuringModels.betapdf

TuringModels.bic

TuringModels.binaryentropy

TuringModels.binomccdf

TuringModels.binomcdf

TuringModels.binominvccdf

TuringModels.binominvcdf

TuringModels.binominvlogccdf

TuringModels.binominvlogcdf

TuringModels.binomlogccdf

TuringModels.binomlogcdf

TuringModels.binomlogpdf

TuringModels.binompdf

TuringModels.blue

TuringModels.boxplot

TuringModels.boxplot!

TuringModels.brush

TuringModels.by

TuringModels.canonform

TuringModels.categorical

TuringModels.categorical!

TuringModels.ccdf

TuringModels.ccolor

TuringModels.cdensity

TuringModels.cdensity!

TuringModels.cdf

TuringModels.center

TuringModels.cf

TuringModels.cgf

TuringModels.cgrad

TuringModels.cgradients

TuringModels.chisqccdf

TuringModels.chisqcdf

TuringModels.chisqinvccdf

TuringModels.chisqinvcdf

TuringModels.chisqinvlogccdf

TuringModels.chisqinvlogcdf

TuringModels.chisqlogccdf

TuringModels.chisqlogcdf

TuringModels.chisqlogpdf

TuringModels.chisqpdf

TuringModels.cie_color_match

TuringModels.circvar

TuringModels.clibraries

TuringModels.clibrary

TuringModels.closeall

TuringModels.coalesce

TuringModels.coef

TuringModels.coefnames

TuringModels.coeftable

TuringModels.color

TuringModels.color_type

TuringModels.coloralpha

TuringModels.colordiff

TuringModels.colormap

TuringModels.colormatch

TuringModels.colwise

TuringModels.combine

TuringModels.comp1

TuringModels.comp2

TuringModels.comp3

TuringModels.compact

TuringModels.competerank

TuringModels.complete_cases

TuringModels.complete_cases!

TuringModels.completecases

TuringModels.component

TuringModels.components

TuringModels.componentwise_logpdf

TuringModels.componentwise_pdf

TuringModels.compress

TuringModels.concentration

TuringModels.confint

TuringModels.consume

TuringModels.contour

TuringModels.contour!

TuringModels.contour3d

TuringModels.contour3d!

TuringModels.contourf

TuringModels.contourf!

TuringModels.convert

TuringModels.coords

TuringModels.cor

TuringModels.cor2cov

TuringModels.corkendall

TuringModels.corner

TuringModels.corner!

TuringModels.cornerplot

TuringModels.cornerplot!

TuringModels.corrplot

TuringModels.corrplot!

TuringModels.corspearman

TuringModels.counteq

TuringModels.countmap

TuringModels.countne

TuringModels.counts

TuringModels.cov

TuringModels.cov2cor

TuringModels.cpu

TuringModels.cquantile

TuringModels.crosscor

TuringModels.crosscor!

TuringModels.crosscov

TuringModels.crosscov!

TuringModels.crossentropy

TuringModels.cumulant

TuringModels.current

TuringModels.curve_points

TuringModels.curves

TuringModels.curves!

TuringModels.cut

TuringModels.cvec

TuringModels.dataset

TuringModels.dataviewer

TuringModels.decompress

TuringModels.default

TuringModels.default_cgrad

TuringModels.delete!

TuringModels.deletecols!

TuringModels.deleterows!

TuringModels.denserank

TuringModels.density

TuringModels.density!

TuringModels.densityplot

TuringModels.describe

TuringModels.deserialize

TuringModels.deuteranopic

TuringModels.deviance

TuringModels.dim

TuringModels.disallowmissing

TuringModels.disallowmissing!

TuringModels.discretediag

TuringModels.distinguishable_colors

TuringModels.diverging_palette

TuringModels.dof

TuringModels.dof_residual

TuringModels.droplevels!

TuringModels.dropmissing

TuringModels.dropmissing!

TuringModels.ea_histogram

TuringModels.ea_histogram!

TuringModels.ecdf

TuringModels.eltypes

TuringModels.elu

TuringModels.enable_stack_copying

TuringModels.entropy

TuringModels.estimate

TuringModels.eval

TuringModels.expected_logdet

TuringModels.f32

TuringModels.f64

TuringModels.failprob

TuringModels.fdistccdf

TuringModels.fdistcdf

TuringModels.fdistinvccdf

TuringModels.fdistinvcdf

TuringModels.fdistinvlogccdf

TuringModels.fdistinvlogcdf

TuringModels.fdistlogccdf

TuringModels.fdistlogcdf

TuringModels.fdistlogpdf

TuringModels.fdistpdf

TuringModels.findat

TuringModels.fit

TuringModels.fit!

TuringModels.fit_map

TuringModels.fit_map!

TuringModels.fit_mle

TuringModels.fit_mle!

TuringModels.fitted

TuringModels.font

TuringModels.fourinvπ

TuringModels.fourπ

TuringModels.frame

TuringModels.freecumulant

TuringModels.fweights

TuringModels.gammaccdf

TuringModels.gammacdf

TuringModels.gammainvccdf

TuringModels.gammainvcdf

TuringModels.gammainvlogccdf

TuringModels.gammainvlogcdf

TuringModels.gammalogccdf

TuringModels.gammalogcdf

TuringModels.gammalogpdf

TuringModels.gammapdf

TuringModels.gamutmax

TuringModels.gamutmin

TuringModels.gelmandiag

TuringModels.generate

TuringModels.generate_t — Function.

generate_t

Utility function to generate all notebooks and chapters from scripts in the scripts directory.

Method

generate_t(sd = script_dict_t)

Required arguments

None, all notebooks/.. and chapters/.. files are regenerated.

generate_t

Generate notebooks and scripts in a single chapter.

Method

generate_t(chapter::AbstractString)

Required arguments

Generate notebooks and scripts in a single chapter, e.g. generate("04")

generate_t

Generate a single notebook and script

Method

generate_t(chapter::AbstractString, file::AbstractString)

Required arguments

Generate notebook and script file in chapter, e.g. generatet("04", "m4.1d.jl") or generatet("04/m4.1t.jl")

TuringModels.genmean

TuringModels.genvar

TuringModels.geomean

TuringModels.get

TuringModels.get_color_palette

TuringModels.getindex

TuringModels.gewekediag

TuringModels.gif

TuringModels.gkldiv

TuringModels.gpu

TuringModels.gr

TuringModels.gradlogpdf

TuringModels.gray

TuringModels.green

TuringModels.grid

TuringModels.groupby

TuringModels.groupedbar

TuringModels.groupedbar!

TuringModels.groupindices

TuringModels.groupvars

TuringModels.gui

TuringModels.halfπ

TuringModels.harmmean

TuringModels.hasfinitesupport

TuringModels.hdf5

TuringModels.head

TuringModels.heatmap

TuringModels.heatmap!

TuringModels.heideldiag

TuringModels.hex

TuringModels.hexbin

TuringModels.hexbin!

TuringModels.hist

TuringModels.histogram

TuringModels.histogram!

TuringModels.histogram2d

TuringModels.histogram2d!

TuringModels.histogramplot

TuringModels.hline

TuringModels.hline!

TuringModels.hspan

TuringModels.hspan!

TuringModels.hyperccdf

TuringModels.hypercdf

TuringModels.hyperinvccdf

TuringModels.hyperinvcdf

TuringModels.hyperinvlogccdf

TuringModels.hyperinvlogcdf

TuringModels.hyperlogccdf

TuringModels.hyperlogcdf

TuringModels.hyperlogpdf

TuringModels.hyperpdf

TuringModels.include

TuringModels.indexmap

TuringModels.indicatormat

TuringModels.informationmatrix

TuringModels.inline

TuringModels.insert!

TuringModels.insertcols!

TuringModels.inspectdr

TuringModels.insupport

TuringModels.inv2π

TuringModels.inv4π

TuringModels.invcov

TuringModels.inverse_rle

TuringModels.invisible

TuringModels.invlogccdf

TuringModels.invlogcdf

TuringModels.invscale

TuringModels.invsoftplus

TuringModels.invsqrt2

TuringModels.invsqrt2π

TuringModels.invπ

TuringModels.iqr

TuringModels.isbounded

TuringModels.isdark

TuringModels.isfitted

TuringModels.isleptokurtic

TuringModels.islinear

TuringModels.islowerbounded

TuringModels.ismesokurtic

TuringModels.ismissing

TuringModels.isordered

TuringModels.isplatykurtic

TuringModels.isprobvec

TuringModels.isupperbounded

TuringModels.iter_segments

TuringModels.kde

TuringModels.keys

TuringModels.kldivergence

TuringModels.kurtosis

TuringModels.leakyrelu

TuringModels.length

TuringModels.levels

TuringModels.levels!

TuringModels.levelsmap

TuringModels.leverage

TuringModels.link

TuringModels.location

TuringModels.location!

TuringModels.log1mexp

TuringModels.log1pexp

TuringModels.log1pmx

TuringModels.log1psq

TuringModels.log2mexp

TuringModels.log2π

TuringModels.log4π

TuringModels.logaddexp

TuringModels.logccdf

TuringModels.logcdf

TuringModels.logdetcov

TuringModels.logexpm1

TuringModels.loghalf

TuringModels.logistic

TuringModels.logit

TuringModels.loglikelihood

TuringModels.logmvgamma

TuringModels.logmxp1

TuringModels.logpdf

TuringModels.logpdf!

TuringModels.logsigmoid

TuringModels.logsoftmax

TuringModels.logsumexp

TuringModels.logtwo

TuringModels.logπ

TuringModels.logσ

TuringModels.lstirling_asym

TuringModels.mad

TuringModels.mad!

TuringModels.map

TuringModels.mapc

TuringModels.mapcols

TuringModels.mapleaves

TuringModels.mapreducec

TuringModels.marginalhist

TuringModels.marginalhist!

TuringModels.maxad

TuringModels.maxpool

TuringModels.mean

TuringModels.mean_and_cov

TuringModels.mean_and_std

TuringModels.mean_and_var

TuringModels.meanad

TuringModels.meandir

TuringModels.meanform

TuringModels.meanlogx

TuringModels.meanplot

TuringModels.meanplot!

TuringModels.meanpool

TuringModels.meanresponse

TuringModels.median

TuringModels.melt

TuringModels.meltdf

TuringModels.merge!

TuringModels.mgf

TuringModels.middle

TuringModels.midpoints

TuringModels.missing

TuringModels.missings

TuringModels.mixeddensity

TuringModels.mixeddensity!

TuringModels.mixeddensityplot

TuringModels.mode

TuringModels.model_response

TuringModels.modelmatrix

TuringModels.modes

TuringModels.moment

TuringModels.mov

TuringModels.mp4

TuringModels.msd

TuringModels.mss

TuringModels.names!

TuringModels.nbetaccdf

TuringModels.nbetacdf

TuringModels.nbetainvccdf

TuringModels.nbetainvcdf

TuringModels.nbetainvlogccdf

TuringModels.nbetainvlogcdf

TuringModels.nbetalogccdf

TuringModels.nbetalogcdf

TuringModels.nbetalogpdf

TuringModels.nbetapdf

TuringModels.nbinomccdf

TuringModels.nbinomcdf

TuringModels.nbinominvccdf

TuringModels.nbinominvcdf

TuringModels.nbinominvlogccdf

TuringModels.nbinominvlogcdf

TuringModels.nbinomlogccdf

TuringModels.nbinomlogcdf

TuringModels.nbinomlogpdf

TuringModels.nbinompdf

TuringModels.ncategories

TuringModels.nchisqccdf

TuringModels.nchisqcdf

TuringModels.nchisqinvccdf

TuringModels.nchisqinvcdf

TuringModels.nchisqinvlogccdf

TuringModels.nchisqinvlogcdf

TuringModels.nchisqlogccdf

TuringModels.nchisqlogcdf

TuringModels.nchisqlogpdf

TuringModels.nchisqpdf

TuringModels.ncol

TuringModels.ncomponents

TuringModels.nfdistccdf

TuringModels.nfdistcdf

TuringModels.nfdistinvccdf

TuringModels.nfdistinvcdf

TuringModels.nfdistinvlogccdf

TuringModels.nfdistinvlogcdf

TuringModels.nfdistlogccdf

TuringModels.nfdistlogcdf

TuringModels.nfdistlogpdf

TuringModels.nfdistpdf

TuringModels.nobs

TuringModels.nonunique

TuringModels.norepeats

TuringModels.normccdf

TuringModels.normcdf

TuringModels.norminvccdf

TuringModels.norminvcdf

TuringModels.norminvlogccdf

TuringModels.norminvlogcdf

TuringModels.normlogccdf

TuringModels.normlogcdf

TuringModels.normlogpdf

TuringModels.normpdf

TuringModels.nquantile

TuringModels.nrow

TuringModels.nsamples

TuringModels.ntdistccdf

TuringModels.ntdistcdf

TuringModels.ntdistinvccdf

TuringModels.ntdistinvcdf

TuringModels.ntdistinvlogccdf

TuringModels.ntdistinvlogcdf

TuringModels.ntdistlogccdf

TuringModels.ntdistlogcdf

TuringModels.ntdistlogpdf

TuringModels.ntdistpdf

TuringModels.ntrials

TuringModels.nullable!

TuringModels.nulldeviance

TuringModels.nullloglikelihood

TuringModels.ohlc

TuringModels.ohlc!

TuringModels.optimize_datetime_ticks

TuringModels.optimize_ticks

TuringModels.order

TuringModels.ordered

TuringModels.ordered!

TuringModels.ordinalrank

TuringModels.pacf

TuringModels.pacf!

TuringModels.palette

TuringModels.param

TuringModels.params

TuringModels.params!

TuringModels.partialcor

TuringModels.partype

TuringModels.passmissing

TuringModels.path3d

TuringModels.path3d!

TuringModels.pdf

TuringModels.percentile

TuringModels.permutecols!

TuringModels.pgfplots

TuringModels.pie

TuringModels.pie!

TuringModels.plot

TuringModels.plot!

TuringModels.plot3d

TuringModels.plot3d!

TuringModels.plot_color

TuringModels.plotarea

TuringModels.plotattr

TuringModels.plotly

TuringModels.plotlyjs

TuringModels.plots_heatmap

TuringModels.plots_heatmap!

TuringModels.png

TuringModels.poisccdf

TuringModels.poiscdf

TuringModels.poisinvccdf

TuringModels.poisinvcdf

TuringModels.poisinvlogccdf

TuringModels.poisinvlogcdf

TuringModels.poislogccdf

TuringModels.poislogcdf

TuringModels.poislogpdf

TuringModels.poispdf

TuringModels.pool

TuringModels.pool!

TuringModels.portfoliocomposition

TuringModels.portfoliocomposition!

TuringModels.predict

TuringModels.predict!

TuringModels.probs

TuringModels.probval

TuringModels.produce

TuringModels.proportionmap

TuringModels.proportions

TuringModels.protanopic

TuringModels.psnr

TuringModels.pweights

TuringModels.pyplot

TuringModels.qqbuild

TuringModels.qqnorm

TuringModels.qqnorm!

TuringModels.qqplot

TuringModels.qqplot!

TuringModels.quantile

TuringModels.quartπ

TuringModels.quiver

TuringModels.quiver!

TuringModels.r2

TuringModels.rafterydiag

TuringModels.rate

TuringModels.readdlm

TuringModels.readtable

TuringModels.recode

TuringModels.recode!

TuringModels.red

TuringModels.reducec

TuringModels.rel_path

TuringModels.rel_path_t — Function.

relpatht

Relative path using the TuringModels src/ directory.

Example to get access to the data subdirectory

rel_path_t("..", "data")

TuringModels.relu

TuringModels.rename

TuringModels.rename!

TuringModels.renyientropy

TuringModels.residuals

TuringModels.response

TuringModels.resume

TuringModels.rgb_string

TuringModels.rgba_string

TuringModels.rle

TuringModels.rmsd

TuringModels.rotate

TuringModels.rotate!

TuringModels.rss

TuringModels.r²

TuringModels.sample

TuringModels.sample!

TuringModels.samplepair

TuringModels.sampler

TuringModels.savefig

TuringModels.scale

TuringModels.scale!

TuringModels.scatter

TuringModels.scatter!

TuringModels.scatter3d

TuringModels.scatter3d!

TuringModels.scatterhist

TuringModels.scatterhist!

TuringModels.scattermat

TuringModels.score

TuringModels.script_dict

TuringModels.script_dict_t

TuringModels.scriptentry

TuringModels.selu

TuringModels.sem

TuringModels.sequential_palette

TuringModels.serialize

TuringModels.set_theme

TuringModels.setadbackend

TuringModels.setadsafe

TuringModels.setchunksize

TuringModels.setindex!

TuringModels.shape

TuringModels.shape_coords

TuringModels.show

TuringModels.showall

TuringModels.showcols

TuringModels.showgradient

TuringModels.showgradient!

TuringModels.showlibrary

TuringModels.showlibrary!

TuringModels.showtheme

TuringModels.showtheme!

TuringModels.sigmoid

TuringModels.skewness

TuringModels.softmax

TuringModels.softmax!

TuringModels.softplus

TuringModels.softsign

TuringModels.span

TuringModels.spy

TuringModels.spy!

TuringModels.sqL2dist

TuringModels.sqmahal

TuringModels.sqmahal!

TuringModels.sqrt2

TuringModels.sqrt2π

TuringModels.sqrt3

TuringModels.sqrt4π

TuringModels.sqrthalfπ

TuringModels.sqrtπ

TuringModels.src_path_t

TuringModels.srdistccdf

TuringModels.srdistcdf

TuringModels.srdistinvccdf

TuringModels.srdistinvcdf

TuringModels.srdistinvlogccdf

TuringModels.srdistinvlogcdf

TuringModels.srdistlogccdf

TuringModels.srdistlogcdf

TuringModels.stack

TuringModels.stackdf

TuringModels.std

TuringModels.stderror

TuringModels.stdlogx

TuringModels.stephist

TuringModels.stephist!

TuringModels.sticks

TuringModels.sticks!

TuringModels.stroke

TuringModels.sub

TuringModels.succprob

TuringModels.suffstats

TuringModels.summarystats

TuringModels.support

TuringModels.surface

TuringModels.surface!

TuringModels.swish

TuringModels.tail

TuringModels.tdistccdf

TuringModels.tdistcdf

TuringModels.tdistinvccdf

TuringModels.tdistinvcdf

TuringModels.tdistinvlogccdf

TuringModels.tdistinvlogcdf

TuringModels.tdistlogccdf

TuringModels.tdistlogcdf

TuringModels.tdistlogpdf

TuringModels.tdistpdf

TuringModels.test_distr

TuringModels.test_examples

TuringModels.test_samples

TuringModels.testhist

TuringModels.testhist!

TuringModels.text

TuringModels.tfill

TuringModels.theme

TuringModels.tiedrank

TuringModels.title!

TuringModels.totalvar

TuringModels.traceplot

TuringModels.traceplot!

TuringModels.transform

TuringModels.translate

TuringModels.translate!

TuringModels.trim

TuringModels.trim!

TuringModels.trimvar

TuringModels.tritanopic

TuringModels.turnprogress

TuringModels.twinx

TuringModels.twoinvπ

TuringModels.twoπ

TuringModels.tzeros

TuringModels.uncompact

TuringModels.unicodeplots

TuringModels.unique!

TuringModels.unstack

TuringModels.values

TuringModels.var

TuringModels.variation

TuringModels.varlogx

TuringModels.vcat

TuringModels.vcov

TuringModels.violin

TuringModels.violin!

TuringModels.vline

TuringModels.vline!

TuringModels.vspan

TuringModels.vspan!

TuringModels.weighted_color_mean

TuringModels.weights

TuringModels.whitebalance

TuringModels.winsor

TuringModels.winsor!

TuringModels.wireframe

TuringModels.wireframe!

TuringModels.with

TuringModels.wmean

TuringModels.wmean!

TuringModels.wmedian

TuringModels.wquantile

TuringModels.wrap

TuringModels.writedlm

TuringModels.writetable

TuringModels.wsample

TuringModels.wsample!

TuringModels.wsum

TuringModels.wsum!

TuringModels.xaxis!

TuringModels.xflip!

TuringModels.xgrid!

TuringModels.xlabel!

TuringModels.xlims

TuringModels.xlims!

TuringModels.xlogx

TuringModels.xlogy

TuringModels.xticks!

TuringModels.xyY

TuringModels.xyYA

TuringModels.yaxis!

TuringModels.yflip!

TuringModels.ygrid!

TuringModels.ylabel!

TuringModels.ylims

TuringModels.ylims!

TuringModels.yticks!

TuringModels.zlims

TuringModels.zlims!

TuringModels.zscore

TuringModels.zscore!

TuringModels.σ