Docstrings

SchattenNorms.E_ — Function.

E(iddim, ρ_rim)

Generates linear map E such that E(ρ) → 1 ⊗ ρ

Note that the dual of this map is F such that F*vec(σ ⊗ ρ) → trace(σ) vec(ρ), in other words, E is the dual of the partial trace.

SchattenNorms.SchattenNorms

SchattenNorms.bra

SchattenNorms.ddist — Function.

ddist(U,V)

Diamond norm distance between two linear CPTP superoperators. Equivalent to dnrom(E-F), but under the assumption that E and F are CPTP supeoperators, but the ddist call is much more accurate and faster due to properties of CPTP superoperators.

See ddistu for a similar implementation of dnorm customized to the difference between unitary operations.

SchattenNorms.ddistu — Function.

ddist(U,V)

Diamond norm distance between two unitary operations. Equivalent to dnorm(liou(U)-liou(V)), under the assumption U and V are unitary matrices. However the ddist call is much more accurate and much faster due to properties of unitary matrices and the corresponding superoperators.

Note: for this particular case, the matrices in question are not the superoperators corresponding to the unitary operation, but rather the unitary operations themselves.

See ddist for a similar implementation of dnorm customized to the difference between CPTP maps.

SchattenNorms.dist_to_neighbour

SchattenNorms.dnorm — Function.

dnorm(L; solver=Convex.getdefaultsolver())

Computes the diamond norm of a linear superoperator L (i.e., a linear transformation of operators). The superoperator must be represented in column major form. In other words, it must be given by a matrix that, when multiplying a vectorized (column major) operator, it should result in the vectorized (column major) representation of the result of the transformation.

SchattenNorms.dnorm2 — Function.

dnorm2(L1,L2)

Computes the diamond norm of a linear superoperator L (i.e., a linear transformation of operators). The superoperator must be represented in column major form. In other words, it must be given by a matrix that, when multiplying a vectorized (column major) operator, it should result in the vectorized (column major) representation of the result of the transformation.

SchattenNorms.dnormcptp — Function.

dnormcptp(L1,L2)

Computes the diamond norm distance between two linear completely positive and trace preserving superoperators L1 and L2 . The superoperators must be represented in column major form.

SchattenNorms.dnormcptp2 — Function.

dnormcptp(L1,L2)

Computes the diamond norm of a linear superoperator L (i.e., a linear transformation of operators). The superoperator must be represented in column major form. In other words, it must be given by a matrix that, when multiplying a vectorized (column major) operator, it should result in the vectorized (column major) representation of the result of the transformation.

SchattenNorms.eval

SchattenNorms.fnorm — Function.

fnorm(m)

Computes the Frobenius norm of a matrix m.

SchattenNorms.include

SchattenNorms.involution — Function.

involution(m)

Permutes the elements of a matrix so that it transforms the column major representation of a linear map L into a matrix C that is positive iff L is completely positive, Hermitian iff L maps (vectorized) Hermitian matrices to Hermitian matrices. In other words, it corresponds to the Choi-Jamiolkoski isomorphism.

SchattenNorms.ket

SchattenNorms.nucnorm — Function.

nucnorm(m)

Computes the nuclear norm of a matrix m.

SchattenNorms.orig_to_line

SchattenNorms.retrϕ — Function.

Compute the trace of the real representation of a complex matrix.

SchattenNorms.snorm — Function.

snorm(m,p=2)

Computes the p-Schatten norm of a matrix m.

SchattenNorms.specnorm — Function.

specnorm(m)

Computes the spectral norm of a matrix m (i.e., the maximum singular value).

SchattenNorms.trnorm — Function.

trnorm(m)

Computes the trace norm of a matrix m.

SchattenNorms.worstfidelity — Function.

worstfidelity(U,V)

For two unitaries U and V, worstfidelity(U,V) is the worst case Jozsa fidelity between U*a and V*a, where a is a complex vector with norm 1. That is, worstfidelity(U,V) is equal to the minimum of |a'*U'*V*a|^2 over all complex vectors a with unit norm.

SchattenNorms.ϕ — Function.

ϕ represents the isomorphism between complex and real matrices. e.g., see Invariant semidefinite programs by Bachoc et al. If two arguments are given, they are taken to be the real and imaginary parts of a complex matrix.

SchattenNorms.ϕi — Function.

Extract the imag part of a complex matrix represented as a real matrix.

SchattenNorms.ϕinv — Function.

Extract the real and imag parts of a complex matrix represented as a real matrix.

SchattenNorms.ϕr — Function.

Extract the real part of a complex matrix represented as a real matrix