Docstrings

Psychro.@polyeval — Macro.

@polyeval(z, p, N)

Evaluate the polynomial $\sum_k c[k] z^{k-1}$ for the coefficients c[1], c[2], ..., c[N]; that is, the coefficients are given in ascending order by power of z. This macro expands to efficient inline code that uses either Horner's method.

julia> p = randn(5)
5-element Array{Float64,1}:
  0.106455 
  0.0351716
 -0.204764 
 -0.87642  
  0.260407 

julia> @polyeval(0.5, p, 5)
-0.020427265391245605

Psychro.AbstractFluid — Type.

Abstract type that represents any kind of fluid

Psychro.AbstractMaterial — Type.

Abstract type that represents any kind of matter

Psychro.Baa — Function.

```Baa(Tk)```

Virial coefficient Bₐₐ of dry air eq 10 [2]

Tk Temperature in K

Output: Bₐₐ in m^3/mol

Psychro.Baw — Function.

```Baw(Tk)```

Cross virial coefficient Baw of saturated vapor eq 15 [2]

Tk Temperature in K

Output: Baw in m^3/mol

Psychro.Blin — Function.

```Blin(Tk)```

Virial coefficient B' saturated vapor eq 15 [1]

Tk Temperature in K

Output: B' in Pa^(-1)

Psychro.Bm — Function.

```Bm(Tk,xv)```

First virial coefficient Bm of moist air of saturated vapor eq 2 [2]

Tk Temperature in K

xv Molar fraction of water vapor in the moist air

Output: Bm in m^3/mol

Psychro.Bww — Function.

```Bww(Tk)```

Virial coefficient Bww of saturated vapor eq 19 [2]

Tk Temperature in K

Output: Bww in m^3/mol

Psychro.Caaa — Function.

```Caa(Tk)```

Second virial coefficient Cₐₐₐ of dry air eq 11 [2].

Tk Temperature in K

Output: Bₐₐ in m^6/mol^2

Psychro.Caaw — Function.

```Caaw(Tk)```

Cross virial coefficient Caaw of saturated vapor eq 16 [2]

Tk Temperature in K

Output: Caaw in m^6/mol^2

Psychro.Caww — Function.

```Caww(Tk)```

Cross virial coefficient Caaw of saturated vapor eq 17 [2]

Tk Temperature in K

Output: Caww in m^6/mol^2

Psychro.Clin — Function.

Second virial coefficient C' saturated vapor eq 16 [1]

Tk Temperature in K

Output: B' in Pa^(-2)

Psychro.Cm — Function.

```Cm(Tk,xv)```

Second virial coefficient Cm of moist air of saturated vapor eq 3 [2]

Tk Temperature in K

xv Molar fraction of water vapor in the moist air

Output: Cm in m^6/mol^2

Psychro.ConvergenceError — Type.

ConvergenceError([msg, val, niter, err])

Exception that is to be thrown when an iterative procedure fails to converge.

Psychro.Cwww — Function.

```Cwww(Tk)```

Second virial coefficient Cwww of saturated vapor eq 20 [2]

Tk Temperature in K

Output: Cwww in m^3/mol

Psychro.Density

Psychro.DewPoint

Psychro.DryAir — Type.

Type used to model dry air

This uses the formulations presente in the ASHRAE handbook Psychrometrics: Theory and Practice, reference [5].

The methods dealing with this type implement the equations of reference [2].

Some thermodynamic properties are implemented as methods. Every method takes as arguments, the type Vaporas first argument and the temperature as the second. IfUnitful` package is used, any unit can be used. On the other hand if no units are provided, the packages assumes SI units all through. In particular:

Temperature in K

Pressure in Pa

Length in m

Quantity of matter in mol

Mass in kg

Energy in J

The following methods are implemented:

volume(DryAir, T, P) specific volume

volumem(DryAir, T, P molar volume

density(DryAir, T, P) density

enthalpy(DryAir, T, P) specific enthalpy

enthalpym(DryAir, T, P) molar enthalpy

entropy(DryAir, T, P) specific entropy

entropym(DryAir, T, P) molar entropy

compressfactor(DryAir, T, P) Compressibility factor.

Psychro.Enthalpy

Psychro.Entropy

Psychro.HumRat

Psychro.L

Psychro.M

Psychro.Ma — Constant.

Molecular weight of air in kg/mol

Psychro.MoistAir — Type.

Type used to model moist air

This uses the formulations presente in the ASHRAE handbook Psychrometrics: Theory and Practice, reference [5].

The temperature range is 173.15 < T < 474.15 K with pressures P < 5 MPa.

The methods dealing with this type implement the equations of references [1-2] to calculate some thermodynamics of real moits air.

Some thermodynamic properties are implemented as methods. Every method takes as arguments, the type MoistAir as first argument, temperature as second argument, a type declaring how humidity is specified, the value of humidity using and pressure. If Unitful package is used, any unit can be used. On the other hand if no units are provided, the packages assumes SI units all through. In particular:

Temperature in K

Pressure in Pa

Length in m

Quantity of matter in mol

Mass in kg

Energy in J

To specify humidity the following types are implemented:

WetBulb for wet bulb temperature, actually adiabatic saturation temperature

DewPoint Dew point temperature

RelHum Relative humidity

HumRat Humidity ratio (kg of vapor / kg of dry air)

SpecHum Specific humidity (kg of vapor / kg of moist air)

MolarFrac Molar fraction of water vapor.

The following methods are implemented:

volume(MoistAir, T, HumType, h, P) specific volume

volumem(MoistAir, T, HumType, h, P molar volume

density(MoistAir, T, HumType, h, P) density

enthalpy(MoistAir, T, HumType, h, P) specific enthalpy

enthalpym(MoistAir, T, HumType, h, P) molar enthalpy

entropy(MoistAir, T, HumType, h, P) specific entropy

entropym(MoistAir, T, HumType, h, P) molar entropy

compressfactor(MoistAir, T, HumType, h, P) Compressibility factor.

It should be noted that the specific enthalpy, entropy and volume are calculated with respect to the mass of dry air. Therefore, 1 J/kg is actually 1 J per kg of dry air.

Psychro.MolarFrac

Psychro.Mv — Constant.

Molecular weight of water in kg/mol

Psychro.Pressure

Psychro.Psychro

Psychro.PsychroProperty

Psychro.Pws — Function.

```Pws(Tk)```

Saturation pressure of vapor pressure over liquid water or ice. This function calls either Pws_l or Pws_s. At a temperature of 273.16 both expressions are almost exactly the same (6 decimal figures).

Tk Temperature in K

Output: Pa

Psychro.Pws_l — Function.

```Pws_l(Tk)```

Saturation pressure of vapor pressure over liquid water. This implements equation 17 from [1].

Tk Temperature in K

Output: Pa

Psychro.Pws_s — Function.

```Pws_s(Tk)```

Saturation pressure of vapor pressure over ice. This implements equation 18 from [1].

Tk Temperature in K

Output: Pa

Psychro.R — Constant.

Universal gas constant kg m^2 /s^2 /K /mol

Psychro.RelHum

Psychro.SpecHum — Type.

Specific Humidity mv / (ma+mv)

Psychro.T0 — Constant.

Melting point of water

Psychro.Temperature

Psychro.ThermodynamicProperty

Psychro.Tws — Function.

```Tws(P)```

Calculates the saturation temperature of water vapor. This function is the inverse of function Pws(T). First an approximation was obtained and then a Newton iteration is used to obtain more accurate data.

P Saturation pressure in Pa

Output: Saturation temperature in K.

Psychro.Vapor — Type.

Type used to model saturated water vapor

This uses the formulations presente in the ASHRAE handbook Psychrometrics: Theory and Practice, reference [5].

The methods dealing with this type implement the equations of reference [1].

Some thermodynamic properties are implemented as methods. Every method takes as arguments, the type DryAir as first argument and the saturation temperature. If Unitful package is used, any unit can be used. On the other hand if no units are provided, the packages assumes SI units all through. In particular:

Temperature in K

Pressure in Pa

Length in m

Quantity of matter in mol

Mass in kg

Energy in J

The following methods are implemented:

volume(Vapor, T) specific volume

volumem(Vapor, T) molar volume

density(Vapor, T) density

enthalpy(Vapor, T) specific enthalpy

enthalpym(Vapor, T) molar enthalpy

entropy(Vapor, T) specific entropy

entropym(Vapor, T) molar entropy

compressfactor(Vapor, T) Compressibility factor.

Psychro.Volume

Psychro.WetBulb

Psychro.Zair — Function.

```Zair(Tk, P, EPS, MAXITER)```

Compressibility factor of dry air.

Tk Temperature in K

P Pressure in Pa

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: Z (nondimensional)

Psychro.Zmoist — Function.

```Zmoist(Tk, P, xv)```

Compressibility factor of moist air.

Tk Temperature in K

P Pressure in Pa

xv molar fraction of water vapor.

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: Z (nondimensional)

Psychro.Zvapor — Function.

Zvapor(Tk)

Compressibility factor of saturated vapor. Eq. 21 reference [1].

Tk Temperature in K

Output: m^3/mol

Psychro.a — Constant.

Coefficients a_i to calculate enthalpy of moist air ref[2]

Psychro.aux_WB

Psychro.calc_W_from_B

Psychro.calcdewpoint

Psychro.calcwetbulb

Psychro.calcz — Function.

calcz(vm, b0, c0, [EPS, [MAXITER, [relax]]])

Calculates the compressibility factor using a virial equation:

z = 1 + b₀/z + c₀/vₘ²

The algorithm initially uses as an initial guess only the b₀ virial coefficient. Then it uses a Newton-Raphson algorithm to compute the compressibility factor

b0 b₀ parameter of the virial equation

c0 c₀ parameter of the virial equation

EPS Convergence criterium

MAXITER Maximim number of iterations

relax Sub-relaxation parameter if convergence is difficult.

Psychro.compressfactor — Function.

compressfactor(DryAir, T, P[, out_unit=u"m^3/kg"])
compressfactor(Vapor, T[, out_unit=u"m^3/kg"])
compressfactor(MoistAir, T, HumidityType, hum, P[, out_unit="m^3/kg"])

Calculates the compressibility factor Z of a gas. In the case of moist air, parameters specifying the humidity shoud be added.

If Unitful units are used, all parameters of the function (except dimensionless) should have units associated. When no units are provided, all parameters should use SI units (and so does the return type). For further information, checkout the doc for Psychro module.

Temperature should be in the range 173.15 K < T < 473.15 and pressure should be below 5 MPa.

Examples

julia> compressfactor(DryAir, 293.15, 101325.0)
1.2044776705391136

julia> compressfactor(DryAir, 20.0u"°C", 1.0u"atm")
1.2044776705391136 kg m^-3

julia> compressfactor(Vapor, 293.15)
0.017308548277505224

julia> compressfactor(Vapor, 20.0u"°C")
0.017308548277505224 kg m^-3

julia> compressfactor(MoistAir, 293.15, RelHum, 0.7, 101325.0)
1.1971436091840653

julia> compressfactor(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm")
1.195819185774941 kg m^-3

Psychro.d — Constant.

Coefficients d_i to calculate enthalpy of moist air ref[2]

Psychro.dBaa — Function.

```dBaa(Tk)```

Derivative of Bₐₐ(T). Calculated from equation 10 [2].

Tk Temperature in K

Output: Bₐₐ in m^3/mol/K

Psychro.dBaw — Function.

```Baw(Tk)```

Derivative of cross virial coefficient Bww of saturated vapor eq 15 [2]

Tk Temperature in K

Output: dBaw/dT in m^3/mol/K

Psychro.dBlin — Function.

```dBlin(Tk)```

Derivative of virial coefficient dB'/dT saturated vapor eq 15 [1]

Tk Temperature in K

Output: B' in Pa^(-1)/K

Psychro.dBm — Function.

```dBm(Tk,xv)```

Derivative of first virial coefficient Bm of moist air of saturated vapor eq 2 [2]

Tk Temperature in K

xv Molar fraction of water vapor in the moist air

Output: dBm/dT in m^3/mol/K

Psychro.dBww — Function.

```dBww(Tk)```

Derivative of virial coefficient dBww/dT of saturated vapor eq 19 [2]

Tk Temperature in K

Output: dBww/dT in m^3/mol/K

Psychro.dCaaa — Function.

```dCaa(Tk)```

Derivative of second virial coefficient Cₐₐₐ of dry air eq 11 [2].

Tk Temperature in K

Output: Bₐₐ in m^6/mol^2/K

Psychro.dCaaw — Function.

```dCaaw(Tk)```

Derivative of cross virial coefficient dCaaw/dT of saturated vapor eq 16 [2]

Tk Temperature in K

Output: dCaaw/dT in m^6/mol^2/K

Psychro.dCaww — Function.

```dCaww(Tk)```

Derivative of cross virial coefficient dCaaw/dT of saturated vapor eq 17 [2]

Tk Temperature in K

Output: dCaww/dT in m^6/mol^2/K

Psychro.dClin — Function.

```dClin(Tk)```

Derivative of Second virial coefficient dC'/dT saturated vapor eq 16 [1]

Tk Temperature in K

Output: B' in Pa^(-2)/K

Psychro.dCm — Function.

```dCm(Tk,xv)```

Derivative of second virial coefficient Cm of moist air of saturated vapor eq 3 [2]

Tk Temperature in K

xv Molar fraction of water vapor in the moist air

Output: dCm/dT in m^6/mol^2/K

Psychro.dCwww — Function.

```dCwww(Tk)```

Derivative of second virial coefficient dCww/dT of saturated vapor eq 20 [2]

Tk Temperature in K

Output: dCwww/dT in m^3/mol/K

Psychro.dPws — Function.

```dPws(Tk)```

Derivative of saturation pressure of vapor pressure over liquid water and ice. This combines the functions dPws_l and dPws_s.

Tk Temperature in K

Output: Pa/K

Psychro.dPws_l — Function.

```dPws_s(Tk)```

Derivative of saturation pressure of vapor pressure over water. This implements the derivative of equation 17 from [1].

Tk Temperature in K

Output: Pa/K

Psychro.dPws_s — Function.

```dPws_s(Tk)```

Derivative of saturation pressure of vapor pressure over ice. This implements the derivative of equation 18 from [1].

Tk Temperature in K

Output: Pa/K

Psychro.density — Function.

density(DryAir, T, P[, out_unit=u"m^3/kg"])
density(Vapor, T[, out_unit=u"m^3/kg"])
density(MoistAir, T, HumidityType, hum, P[, out_unit="m^3/kg"])

Calculates the density volume of a gas. In the case of moist air, parameters specifying the humidity shoud be added.

If Unitful units are used, all parameters of the function (except dimensionless) should have units associated. When no units are provided, all parameters should use SI units (and so does the return type). For further information, checkout the doc for Psychro module.

Temperature should be in the range 173.15 K < T < 473.15 and pressure should be below 5 MPa.

Examples

julia> density(DryAir, 293.15, 101325.0)
1.2044776705391136

julia> density(DryAir, 20.0u"°C", 1.0u"atm")
1.2044776705391136 kg m^-3

julia> density(Vapor, 293.15)
0.017308548277505224

julia> density(Vapor, 20.0u"°C")
0.017308548277505224 kg m^-3

julia> density(MoistAir, 293.15, RelHum, 0.7, 101325.0)
1.1971436091840653

julia> density(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm")
1.195819185774941 kg m^-3

julia> density(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm", u"lb/inch^3")
4.3201708903793606e-5 in^-3 lb

Psychro.densitywater — Function.

densitywater(Tk)

Density of saturated water. Equation 5 from ref. [1].

Tk Temperature in K.

Output: density kg/m^3

Psychro.dewpoint — Function.

Thermodynamic properties of moist air, dry air and saturated water vapor.

The methods listed below calculate thermodynamic properties of moist air:

volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
density(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpy(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpym(MoistAir, T, HumidityType, humidity, P[, outunit])
entropy(MoistAir, T, HumidityType, humidity, P[, outunit])
entropym(MoistAir, T, HumidityType, humidity, P[, outunit])
compressfactor(MoistAir, T, HumidityType, humidity, P[, outunit])
dewpoint(MoistAir, T, HumidityType, humidity, P[, outunit]) 
wetbulb(MoistAir, T, HumidityType, humidity, P[, outunit]) 
humrat(MoistAir, T, HumidityType, humidity, P) 
relhum(MoistAir, T, HumidityType, humidity, P) 
humrat(MoistAir, T, HumidityType, humidity, P) 
spechum(MoistAir, T, HumidityType, humidity, P) 
molarfrac(MoistAir, T, HumidityType, humidity, P)

The methods listed above calculate the following thermodynamic properties of moist air:

volume Specific volume

volumem Molar volume

density Density

enthalpy Specific enthalpy

enthalpym Molar enthalpy

entropy Specific entropy

entropym Molar entropy

compressfactor Compressibility factor Z

dewpoint Dew point temperature

wetbulb Adiabatic saturation temperature

humrat Humidity ratio

relhum Relative humidity

spechum Specific humidity

molarfrac Molar fraction of water vapor

The humidity is specified using two parameters:

How the humidity is specified

The actual value of humidity

The following types are used to characterize the humidity.

WetBulb for wet bulb temperature, actually adiabatic saturation temperature

DewPoint Dew point temperature

RelHum Relative humidity

HumRat Humidity ratio (kg of vapor / kg of dry air)

SpecHum Specific humidity (kg of vapor / kg of moist air)

MolarFrac Molar fraction of water vapor.

Examples

julia> volume(MoistAir, 293.15, WetBulb, 291.15, 101325.0)
0.8464079202783964

julia> volume(MoistAir, 293.15, DewPoint, 291.15, 101325.0)
0.8475219875187474

julia> volume(MoistAir, 293.15, RelHum, 0.7, 101325.0)
0.843889817602806

julia> volume(MoistAir, 20.0u"°C", DewPoint, 60.0u"°F", 1.0u"atm")
0.8449934929585231 kg^-1 m^3

julia> volumem(MoistAir, 293.15, RelHum, 0.5, 93000.0)
0.026199080086890276

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1.0976075893895811 kg m^-3

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"lb/inch^3")
3.965358988338535e-5 in^-3 lb

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> enthalpy(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
50667.43014746832 kg^-1 J

julia> enthalpym(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1439.6551689935861 J mol^-1

julia> compressfactor(MoistAir, -90.0u"°C", RelHum, 0.01, 4.5u"MPa")
0.8552758629097985

julia> wetbulb(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
17.0 °C

julia> dewpoint(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
15.475836053510477 °C

julia> humrat(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.012032930694441925

julia> relhum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.7517801524436909

julia> spechum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.011889860823189923

Psychro.dimless

Psychro.efactor

Psychro.efactor2 — Function.

```e_factor(Tk,P)```

Enhancement factor of moist air. Uses an iterative procedure to calculate the Enhancement factor defined in equation 18 [2]

Psychro.enthalpy — Function.

enthalpy(DryAir, T, P[, out_unit=u"m^3/kg"])
enthalpy(Vapor, T[, out_unit=u"m^3/kg"])
enthalpy(MoistAir, T, HumidityType, hum, P[, out_unit="m^3/kg"])

Calculates the specific enthalpy of a gas. In the case of moist air, parameters specifying the humidity shoud be added and the the property is based on the mass of dry air not mass of the gas (as is the case for DryAir or Vapor).

If Unitful units are used, all parameters of the function (except dimensionless) should have units associated. When no units are provided, all parameters should use SI units (and so does the return type). For further information, checkout the doc for Psychro module.

Temperature should be in the range 173.15 K < T < 473.15 and pressure should be below 5 MPa.

Examples

julia> enthalpy(DryAir, 293.15, 101325.0)
20121.2722813185

julia> enthalpy(DryAir, 20.0u"°C", 1.0u"atm")
20121.2722813185 kg^-1 J

julia> enthalpy(Vapor, 293.15)
2.5374003352412493e6

julia> enthalpy(Vapor, 20.0u"°C")
2.5374003352412493e6 kg^-1 J

julia> enthalpy(MoistAir, 293.15, RelHum, 0.7, 101325.0)
46142.773129687484

julia> enthalpy(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm")
50944.84575377501 kg^-1 J

julia> enthalpy(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm", u"J/lb")
23108.193324739244 J lb^-1

Psychro.enthalpyair — Function.

enthalpyair(Tk)

Specific enthalpy of dry air. Equation 13 of reference [2]

Tk Temperature in K

P Pressure in Pa

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: m^3/kg

Psychro.enthalpyice — Function.

enthalpyice(Tk)

Specific enthalpy of saturated ice. Equation 3 of ref. [1].

Tk Temperature in K

Output: J/kg

Psychro.enthalpym — Function.

enthalpym(DryAir, T, P[, out_unit=u"m^3/kg"])
enthalpym(Vapor, T[, out_unit=u"m^3/kg"])
enthalpym(MoistAir, T, HumidityType, hum, P[, out_unit="m^3/kg"])

Calculates the molar enthalpy of a gas. In the case of moist air, parameters specifying the humidity shoud be added.

If Unitful units are used, all parameters of the function (except dimensionless) should have units associated. When no units are provided, all parameters should use SI units (and so does the return type). For further information, checkout the doc for Psychro module.

Temperature should be in the range 173.15 K < T < 473.15 and pressure should be below 5 MPa.

Examples

julia> enthalpym(DryAir, 293.15, 101325.0)
582.7824697199684

julia> enthalpym(DryAir, 20.0u"°C", 1.0u"atm")
582.7824697199684 J mol^-1

julia> enthalpy(Vapor, 293.15)
2.5374003352412493e6

julia> enthalpym(Vapor, 20.0u"°C")
45711.977511464975 J mol^-1

julia> enthalpym(MoistAir, 293.15, RelHum, 0.7, 101325.0)
1314.7744549269437

julia> enthalpym(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm")
1447.2684837312868 J mol^-1

julia> enthalpym(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm", u"kJ/kmol")
1447.2684837312868 kJ kmol^-1

Psychro.enthalpymoist — Function.

enthalpymoist(Tk, P, xv, EPS=1e-8, MAXITER=100)

Specific enthalphy of moist air defined as enthalpy per kg of dry air.

Tk Temperature in K

P Pressure in Pa

xv molar fraction of water vapor.

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: J/kg of dry air

Psychro.enthalpyvapor — Function.

enthalpyvapor(Tk)

Specific enthalpy of saturated water vapor. Equation 19 of ref. [1].

Equation 19 ha a problem since the terms using the virial coefficients are actually molar enthalpies not specific. Dividing these terms by Mv solves any issues and the equations match the table in the appendix!

Tk Temperature in K

Output: J/kg

Psychro.enthalpywater — Function.

enthalpywater(Tk)

Specific enthalpy of saturated water. Equations 6-11 of ref. [1].

Tk Temperature in K

Output: J/kg

Psychro.enthalpywi — Function.

enthalpywi(Tk)

Specific enthalpy of saturated water in condensed phase (ice or water).

Tk Temperature in K

Output: J/kg

Psychro.entropy — Function.

entropy(DryAir, T, P[, out_unit=u"m^3/kg"])
entropy(Vapor, T[, out_unit=u"m^3/kg"])
entropy(MoistAir, T, HumidityType, hum, P[, out_unit="m^3/kg"])

Calculates the specific entropy of a gas. In the case of moist air, parameters specifying the humidity shoud be added and the the property is based on the mass of dry air not mass of the gas (as is the case for DryAir or Vapor).

If Unitful units are used, all parameters of the function (except dimensionless) should have units associated. When no units are provided, all parameters should use SI units (and so does the return type). For further information, checkout the doc for Psychro module.

Temperature should be in the range 173.15 K < T < 473.15 and pressure should be below 5 MPa.

Examples

julia> entropy(DryAir, 293.15, 101325.0)
71.09262577195514

julia> entropy(DryAir, 20.0u"°C", 1.0u"atm")
71.09262577195514 kg^-1 J K^-1

julia> entropy(Vapor, 293.15)
8665.873003613997

julia> entropy(Vapor, 20.0u"°C")
8665.873003613997 kg^-1 J K^-1

julia> entropy(MoistAir, 293.15, RelHum, 0.7, 101325.0)
166.33538729355692

julia> entropy(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm")
182.97140931650105 kg^-1 J K^-1

julia> entropy(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm", u"kJ/lb/°F")
0.04610801955228433 °F^-1 kJ lb^-1

Psychro.entropyair — Function.

entropyair(Tk)

Specific entropy of dry air. Equation 14 of reference [2]

Tk Temperature in K

P Pressure in Pa

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: m^3/(kg.K)

Psychro.entropym — Function.

entropym(DryAir, T, P[, out_unit=u"m^3/kg"])
entropym(Vapor, T[, out_unit=u"m^3/kg"])
entropym(MoistAir, T, HumidityType, hum, P[, out_unit="m^3/kg"])

Calculates the molar entropy of a gas. In the case of moist air, parameters specifying the humidity shoud be added.

If Unitful units are used, all parameters of the function (except dimensionless) should have units associated. When no units are provided, all parameters should use SI units (and so does the return type). For further information, checkout the doc for Psychro module.

Temperature should be in the range 173.15 K < T < 473.15 and pressure should be below 5 MPa.

Examples

julia> entropym(DryAir, 293.15, 101325.0)
2.0590912665460226

julia> entropym(DryAir, 20.0u"°C", 1.0u"atm")
2.0590912665460226 J K^-1 mol^-1

julia> entropy(Vapor, 293.15)
8665.873003613997

julia> entropym(Vapor, 20.0u"°C")
156.11812860454717 J K^-1 mol^-1

julia> entropym(MoistAir, 293.15, RelHum, 0.7, 101325.0)
4.739496639035868

julia> entropym(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm")
5.197949865380578 J K^-1 mol^-1

julia> entropym(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm", u"kJ/kmol/°F")
2.8877499252114327 °F^-1 kJ kmol^-1

Psychro.entropymoist — Function.

entropymoist(Tk, P, xv, EPS=1e-8, MAXITER=100)

Specific entropy of moist air defined as entropy per kg of dry air.

Tk Temperature in K

P Pressure in Pa

xv molar fraction of water vapor.

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: J/kg/K of dry air

Psychro.entropyvapor — Function.

entropyvapor(Tk)

Specific entropy of saturated vapor. Eq. 20 reference [1].

Tk Temperature in K

Output: J/kg/K

Psychro.eval

Psychro.g — Constant.

Coefficients g_i to calculate enthalpy of moist air ref[2]

Psychro.gg

Psychro.henryk — Function.

henryk(Tk)

Henry's coefficient for solubility of air in water. Implements equations 24 and 25 from ref. [2]. See reference [3] Valid for temperatures between 273.15 K and 500 K

Tk Temperature in K

Output: k in atm/mol fract * 10^(-4)

Psychro.henryk_N2 — Function.

henryk_N2(Tk)

Henry's coefficient for solubility of N2 in water. Implements equation 23 from ref. [2]. See reference [3] Valid for temperatures between 273.15 K and 500 K

Tk Temperature in K

Output: k in atm/mol * 10^(-4)

Psychro.henryk_O2 — Function.

henryk_O2(Tk)

Henry's coefficient for solubility of O2 in water. Implements equation 23 from ref. [2]. See reference [3] Valid for temperatures between 273.15 K and 500 K

Tk Temperature in K

Output: k in atm/mol * 10^(-4)

Psychro.hh

Psychro.humrat — Function.

Thermodynamic properties of moist air, dry air and saturated water vapor.

The methods listed below calculate thermodynamic properties of moist air:

volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
density(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpy(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpym(MoistAir, T, HumidityType, humidity, P[, outunit])
entropy(MoistAir, T, HumidityType, humidity, P[, outunit])
entropym(MoistAir, T, HumidityType, humidity, P[, outunit])
compressfactor(MoistAir, T, HumidityType, humidity, P[, outunit])
dewpoint(MoistAir, T, HumidityType, humidity, P[, outunit]) 
wetbulb(MoistAir, T, HumidityType, humidity, P[, outunit]) 
humrat(MoistAir, T, HumidityType, humidity, P) 
relhum(MoistAir, T, HumidityType, humidity, P) 
humrat(MoistAir, T, HumidityType, humidity, P) 
spechum(MoistAir, T, HumidityType, humidity, P) 
molarfrac(MoistAir, T, HumidityType, humidity, P)

The methods listed above calculate the following thermodynamic properties of moist air:

volume Specific volume

volumem Molar volume

density Density

enthalpy Specific enthalpy

enthalpym Molar enthalpy

entropy Specific entropy

entropym Molar entropy

compressfactor Compressibility factor Z

dewpoint Dew point temperature

wetbulb Adiabatic saturation temperature

humrat Humidity ratio

relhum Relative humidity

spechum Specific humidity

molarfrac Molar fraction of water vapor

The humidity is specified using two parameters:

How the humidity is specified

The actual value of humidity

The following types are used to characterize the humidity.

WetBulb for wet bulb temperature, actually adiabatic saturation temperature

DewPoint Dew point temperature

RelHum Relative humidity

HumRat Humidity ratio (kg of vapor / kg of dry air)

SpecHum Specific humidity (kg of vapor / kg of moist air)

MolarFrac Molar fraction of water vapor.

Examples

julia> volume(MoistAir, 293.15, WetBulb, 291.15, 101325.0)
0.8464079202783964

julia> volume(MoistAir, 293.15, DewPoint, 291.15, 101325.0)
0.8475219875187474

julia> volume(MoistAir, 293.15, RelHum, 0.7, 101325.0)
0.843889817602806

julia> volume(MoistAir, 20.0u"°C", DewPoint, 60.0u"°F", 1.0u"atm")
0.8449934929585231 kg^-1 m^3

julia> volumem(MoistAir, 293.15, RelHum, 0.5, 93000.0)
0.026199080086890276

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1.0976075893895811 kg m^-3

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"lb/inch^3")
3.965358988338535e-5 in^-3 lb

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> enthalpy(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
50667.43014746832 kg^-1 J

julia> enthalpym(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1439.6551689935861 J mol^-1

julia> compressfactor(MoistAir, -90.0u"°C", RelHum, 0.01, 4.5u"MPa")
0.8552758629097985

julia> wetbulb(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
17.0 °C

julia> dewpoint(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
15.475836053510477 °C

julia> humrat(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.012032930694441925

julia> relhum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.7517801524436909

julia> spechum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.011889860823189923

Psychro.include

Psychro.k — Constant.

Coefficients k_i to calculate enthalpy of moist air ref[2]

Psychro.kappa_f

Psychro.kappa_l — Function.

kappa_l(Tk)

Isothermal compressibility of saturated liquid water. Equation 21, ref. [2]. Range of applicability 0°C - 150°C but can be extended up to 200°C More details in reference [4].

Tk Temperature in K

Output: κ in 1/Pa

Psychro.kappa_s — Function.

kappa_s(Tk)

Isothermal compressibility of ice. Equation 22, ref. [2].

Tk Temperature in K

Output: κ in 1/Pa

Psychro.lnf

Psychro.lnf2 — Function.

```lnf(Tk, P, xas)```

Auxiliary function used to calculate the enhancement factor. Actually, this function implements the RHS of eq. 18 of [2].

Psychro.mm

Psychro.molarenthalpyair — Function.

molarenthalpyair(Tk)

Molar enthalpy of dry air. Equation 13 of reference [2]

Tk Temperature in K

P Pressure in Pa

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: m^3/mol

Psychro.molarenthalpymoist — Function.

molarenthalpymoist(Tk, P, xv, EPS=1e-8, MAXITER=100)

Molar enthalphy of moist air defined as enthalpy dry air.

Tk Temperature in K

P Pressure in Pa

xv molar fraction of water vapor.

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: J/mol

Psychro.molarentropyair — Function.

molarentropyair(Tk)

Specific entropy of dry air. Equation 14 of reference [2]

Tk Temperature in K

P Pressure in Pa

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: m^3/(kg.K)

Psychro.molarentropymoist — Function.

molarentropymoist(Tk, P, xv, EPS=1e-8, MAXITER=100)

Molar entropy of moist air defined as entropy per kg of dry air.

Tk Temperature in K

P Pressure in Pa

xv molar fraction of water vapor.

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: J/mol/K of dry air

Psychro.molarentropyvapor — Function.

molarentropyvapor(Tk)

Molar entropy of saturated vapor. Eq. 20 reference [1].

Tk Temperature in K

Output: J/mol/K

Psychro.molarfrac — Function.

Thermodynamic properties of moist air, dry air and saturated water vapor.

The methods listed below calculate thermodynamic properties of moist air:

volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
density(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpy(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpym(MoistAir, T, HumidityType, humidity, P[, outunit])
entropy(MoistAir, T, HumidityType, humidity, P[, outunit])
entropym(MoistAir, T, HumidityType, humidity, P[, outunit])
compressfactor(MoistAir, T, HumidityType, humidity, P[, outunit])
dewpoint(MoistAir, T, HumidityType, humidity, P[, outunit]) 
wetbulb(MoistAir, T, HumidityType, humidity, P[, outunit]) 
humrat(MoistAir, T, HumidityType, humidity, P) 
relhum(MoistAir, T, HumidityType, humidity, P) 
humrat(MoistAir, T, HumidityType, humidity, P) 
spechum(MoistAir, T, HumidityType, humidity, P) 
molarfrac(MoistAir, T, HumidityType, humidity, P)

The methods listed above calculate the following thermodynamic properties of moist air:

volume Specific volume

volumem Molar volume

density Density

enthalpy Specific enthalpy

enthalpym Molar enthalpy

entropy Specific entropy

entropym Molar entropy

compressfactor Compressibility factor Z

dewpoint Dew point temperature

wetbulb Adiabatic saturation temperature

humrat Humidity ratio

relhum Relative humidity

spechum Specific humidity

molarfrac Molar fraction of water vapor

The humidity is specified using two parameters:

How the humidity is specified

The actual value of humidity

The following types are used to characterize the humidity.

WetBulb for wet bulb temperature, actually adiabatic saturation temperature

DewPoint Dew point temperature

RelHum Relative humidity

HumRat Humidity ratio (kg of vapor / kg of dry air)

SpecHum Specific humidity (kg of vapor / kg of moist air)

MolarFrac Molar fraction of water vapor.

Examples

julia> volume(MoistAir, 293.15, WetBulb, 291.15, 101325.0)
0.8464079202783964

julia> volume(MoistAir, 293.15, DewPoint, 291.15, 101325.0)
0.8475219875187474

julia> volume(MoistAir, 293.15, RelHum, 0.7, 101325.0)
0.843889817602806

julia> volume(MoistAir, 20.0u"°C", DewPoint, 60.0u"°F", 1.0u"atm")
0.8449934929585231 kg^-1 m^3

julia> volumem(MoistAir, 293.15, RelHum, 0.5, 93000.0)
0.026199080086890276

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1.0976075893895811 kg m^-3

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"lb/inch^3")
3.965358988338535e-5 in^-3 lb

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> enthalpy(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
50667.43014746832 kg^-1 J

julia> enthalpym(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1439.6551689935861 J mol^-1

julia> compressfactor(MoistAir, -90.0u"°C", RelHum, 0.01, 4.5u"MPa")
0.8552758629097985

julia> wetbulb(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
17.0 °C

julia> dewpoint(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
15.475836053510477 °C

julia> humrat(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.012032930694441925

julia> relhum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.7517801524436909

julia> spechum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.011889860823189923

Psychro.molarfracmoist_sat — Function.

molarfracmoist_sat(Tk, P)

Molar fraction of water vapor in saturated moist air

Tk Temperature in K

P Pressure in Pa

Output: molar fraction

Psychro.molarvolumeair — Function.

molarvolumeair(Tk)

Molar volume of dry air. Equation 12 of reference [1]. This function requires iteration to compute the volume.

Tk Temperature in K

P Pressure in Pa

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: molar volume in m^3/mol

Psychro.molarvolumemoist — Function.

```molarvolumemoist```

Molar volume of moist air given the molar fraction of water vapor. Assumes a real gas.

Tk Temperature in K

P Pressure in Pa

xv Molar fraction of water vapor

Output: molar volume in m^3/mol

Psychro.molarvolvapor — Function.

volumevapor(Tk)

Molar volume of saturated vapor. Eq. 21 reference [1].

Tk Temperature in K

Output: m^3/mol

Psychro.ndim

Psychro.perc

Psychro.relhum — Function.

Thermodynamic properties of moist air, dry air and saturated water vapor.

The methods listed below calculate thermodynamic properties of moist air:

volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
density(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpy(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpym(MoistAir, T, HumidityType, humidity, P[, outunit])
entropy(MoistAir, T, HumidityType, humidity, P[, outunit])
entropym(MoistAir, T, HumidityType, humidity, P[, outunit])
compressfactor(MoistAir, T, HumidityType, humidity, P[, outunit])
dewpoint(MoistAir, T, HumidityType, humidity, P[, outunit]) 
wetbulb(MoistAir, T, HumidityType, humidity, P[, outunit]) 
humrat(MoistAir, T, HumidityType, humidity, P) 
relhum(MoistAir, T, HumidityType, humidity, P) 
humrat(MoistAir, T, HumidityType, humidity, P) 
spechum(MoistAir, T, HumidityType, humidity, P) 
molarfrac(MoistAir, T, HumidityType, humidity, P)

The methods listed above calculate the following thermodynamic properties of moist air:

volume Specific volume

volumem Molar volume

density Density

enthalpy Specific enthalpy

enthalpym Molar enthalpy

entropy Specific entropy

entropym Molar entropy

compressfactor Compressibility factor Z

dewpoint Dew point temperature

wetbulb Adiabatic saturation temperature

humrat Humidity ratio

relhum Relative humidity

spechum Specific humidity

molarfrac Molar fraction of water vapor

The humidity is specified using two parameters:

How the humidity is specified

The actual value of humidity

The following types are used to characterize the humidity.

WetBulb for wet bulb temperature, actually adiabatic saturation temperature

DewPoint Dew point temperature

RelHum Relative humidity

HumRat Humidity ratio (kg of vapor / kg of dry air)

SpecHum Specific humidity (kg of vapor / kg of moist air)

MolarFrac Molar fraction of water vapor.

Examples

julia> volume(MoistAir, 293.15, WetBulb, 291.15, 101325.0)
0.8464079202783964

julia> volume(MoistAir, 293.15, DewPoint, 291.15, 101325.0)
0.8475219875187474

julia> volume(MoistAir, 293.15, RelHum, 0.7, 101325.0)
0.843889817602806

julia> volume(MoistAir, 20.0u"°C", DewPoint, 60.0u"°F", 1.0u"atm")
0.8449934929585231 kg^-1 m^3

julia> volumem(MoistAir, 293.15, RelHum, 0.5, 93000.0)
0.026199080086890276

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1.0976075893895811 kg m^-3

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"lb/inch^3")
3.965358988338535e-5 in^-3 lb

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> enthalpy(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
50667.43014746832 kg^-1 J

julia> enthalpym(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1439.6551689935861 J mol^-1

julia> compressfactor(MoistAir, -90.0u"°C", RelHum, 0.01, 4.5u"MPa")
0.8552758629097985

julia> wetbulb(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
17.0 °C

julia> dewpoint(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
15.475836053510477 °C

julia> humrat(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.012032930694441925

julia> relhum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.7517801524436909

julia> spechum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.011889860823189923

Psychro.spechum — Function.

Thermodynamic properties of moist air, dry air and saturated water vapor.

The methods listed below calculate thermodynamic properties of moist air:

volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
density(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpy(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpym(MoistAir, T, HumidityType, humidity, P[, outunit])
entropy(MoistAir, T, HumidityType, humidity, P[, outunit])
entropym(MoistAir, T, HumidityType, humidity, P[, outunit])
compressfactor(MoistAir, T, HumidityType, humidity, P[, outunit])
dewpoint(MoistAir, T, HumidityType, humidity, P[, outunit]) 
wetbulb(MoistAir, T, HumidityType, humidity, P[, outunit]) 
humrat(MoistAir, T, HumidityType, humidity, P) 
relhum(MoistAir, T, HumidityType, humidity, P) 
humrat(MoistAir, T, HumidityType, humidity, P) 
spechum(MoistAir, T, HumidityType, humidity, P) 
molarfrac(MoistAir, T, HumidityType, humidity, P)

The methods listed above calculate the following thermodynamic properties of moist air:

volume Specific volume

volumem Molar volume

density Density

enthalpy Specific enthalpy

enthalpym Molar enthalpy

entropy Specific entropy

entropym Molar entropy

compressfactor Compressibility factor Z

dewpoint Dew point temperature

wetbulb Adiabatic saturation temperature

humrat Humidity ratio

relhum Relative humidity

spechum Specific humidity

molarfrac Molar fraction of water vapor

The humidity is specified using two parameters:

How the humidity is specified

The actual value of humidity

The following types are used to characterize the humidity.

WetBulb for wet bulb temperature, actually adiabatic saturation temperature

DewPoint Dew point temperature

RelHum Relative humidity

HumRat Humidity ratio (kg of vapor / kg of dry air)

SpecHum Specific humidity (kg of vapor / kg of moist air)

MolarFrac Molar fraction of water vapor.

Examples

julia> volume(MoistAir, 293.15, WetBulb, 291.15, 101325.0)
0.8464079202783964

julia> volume(MoistAir, 293.15, DewPoint, 291.15, 101325.0)
0.8475219875187474

julia> volume(MoistAir, 293.15, RelHum, 0.7, 101325.0)
0.843889817602806

julia> volume(MoistAir, 20.0u"°C", DewPoint, 60.0u"°F", 1.0u"atm")
0.8449934929585231 kg^-1 m^3

julia> volumem(MoistAir, 293.15, RelHum, 0.5, 93000.0)
0.026199080086890276

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1.0976075893895811 kg m^-3

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"lb/inch^3")
3.965358988338535e-5 in^-3 lb

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> enthalpy(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
50667.43014746832 kg^-1 J

julia> enthalpym(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1439.6551689935861 J mol^-1

julia> compressfactor(MoistAir, -90.0u"°C", RelHum, 0.01, 4.5u"MPa")
0.8552758629097985

julia> wetbulb(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
17.0 °C

julia> dewpoint(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
15.475836053510477 °C

julia> humrat(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.012032930694441925

julia> relhum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.7517801524436909

julia> spechum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.011889860823189923

Psychro.uD

Psychro.uH

Psychro.uP

Psychro.uS

Psychro.uT

Psychro.uV

Psychro.umD

Psychro.umH

Psychro.umS

Psychro.umV

Psychro.val

Psychro.volume — Function.

volume(DryAir, T, P[, out_unit=u"m^3/kg"])
volume(Vapor, T[, out_unit=u"m^3/kg"])
volume(MoistAir, T, HumidityType, hum, P[, out_unit="m^3/kg"])

Calculates the specific volume of a gas. In the case of moist air, parameters specifying the humidity shoud be added and the the property is based on the mass of dry air not mass of the gas (as is the case for DryAir or Vapor).

If Unitful units are used, all parameters of the function (except dimensionless) should have units associated. When no units are provided, all parameters should use SI units (and so does the return type). For further information, checkout the doc for Psychro module.

Temperature should be in the range 173.15 K < T < 473.15 and pressure should be below 5 MPa.

Examples

julia> volume(DryAir, 293.15, 101325.0)
0.8302353995092402

julia> volume(DryAir, 20.0u"°C", 1.0u"atm")
0.8302353995092402 kg^-1 m^3

julia> volume(Vapor, 293.15)
57.77492045936827

julia> volume(Vapor, 20.0u"°C")
57.77492045936827 kg^-1 m^3

julia> volume(MoistAir, 293.15, RelHum, 0.7, 101325.0)
0.843889817602806

julia> volume(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm")
0.8464079202783964 kg^-1 m^3

julia> volume(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm", u"inch^3/lb")
23428.490579267214 in^3 lb^-1

Psychro.volumeair — Function.

volumeair

Specific volume of dry air. Uses molarvolumeair.

Tk Temperature in K

P Pressure in Pa

EPS: Acceptable error

MAXITER: Maixmum number of iterations

Output: molar volume in m^3/kg

Psychro.volumeice — Function.

volumeice(Tk)

Specific volume of saturated ice. Equation 2 from ref. [1].

Tk Temperature in K.

Output: specific volume in m^3/kg

Psychro.volumem — Function.

volumem(DryAir, T, P[, out_unit=u"m^3/kg"])
volumem(Vapor, T[, out_unit=u"m^3/kg"])
volumem(MoistAir, T, HumidityType, hum, P[, out_unit="m^3/kg"])

Calculates the molar volume of a gas. In the case of moist air, parameters specifying the humidity shoud be added.

If Unitful units are used, all parameters of the function (except dimensionless) should have units associated. When no units are provided, all parameters should use SI units (and so does the return type). For further information, checkout the doc for Psychro module.

Temperature should be in the range 173.15 K < T < 473.15 and pressure should be below 5 MPa.

Examples

julia> volumem(DryAir, 293.15, 101325.0)
0.024046522993685877

julia> volumem(DryAir, 20.0u"°C", 1.0u"atm")
0.024046522993685877 m^3 mol^-1

julia> volumem(Vapor, 293.15)
1.040831369053248

julia> volumem(Vapor, 20.0u"°C")
1.040831369053248 m^3 mol^-1

julia> volumem(MoistAir, 293.15, RelHum, 0.7, 101325.0)
0.02404547233948706

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm")
0.024045209859305458 m^3 mol^-1

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 18.0u"°C", 1.0u"atm", u"inch^3/mol")
1467.32873315839 in^3 mol^-1

Psychro.volumemoist — Function.

```volumemoist```

Specific volume of moist air given the molar fraction of water vapor. Assumes a real gas. The specific volume is defined as the volume occupied by a mixture of dry air and water vapor per kg of dry air.

Tk Temperature in K

P Pressure in Pa

xv Molar fraction of water vapor

Output: molar volume in m^3/kg of dry air.

Psychro.volumevapor — Function.

volumevapor(Tk)

Specific volume of saturated vapor. Eq. 21 reference [1].

Tk Temperature in K

Output: m^3/kg

Psychro.volumewater — Function.

volumewater(Tk)

Specific volume of saturated water. Inverse of equation 5 from ref. [1].

Tk Temperature in K.

Output: specific volume in m^3/kg

Psychro.wetbulb — Function.

Thermodynamic properties of moist air, dry air and saturated water vapor.

The methods listed below calculate thermodynamic properties of moist air:

volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
volume(MoistAir, T, HumidityType, humidity, P[, outunit]) 
density(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpy(MoistAir, T, HumidityType, humidity, P[, outunit])
enthalpym(MoistAir, T, HumidityType, humidity, P[, outunit])
entropy(MoistAir, T, HumidityType, humidity, P[, outunit])
entropym(MoistAir, T, HumidityType, humidity, P[, outunit])
compressfactor(MoistAir, T, HumidityType, humidity, P[, outunit])
dewpoint(MoistAir, T, HumidityType, humidity, P[, outunit]) 
wetbulb(MoistAir, T, HumidityType, humidity, P[, outunit]) 
humrat(MoistAir, T, HumidityType, humidity, P) 
relhum(MoistAir, T, HumidityType, humidity, P) 
humrat(MoistAir, T, HumidityType, humidity, P) 
spechum(MoistAir, T, HumidityType, humidity, P) 
molarfrac(MoistAir, T, HumidityType, humidity, P)

The methods listed above calculate the following thermodynamic properties of moist air:

volume Specific volume

volumem Molar volume

density Density

enthalpy Specific enthalpy

enthalpym Molar enthalpy

entropy Specific entropy

entropym Molar entropy

compressfactor Compressibility factor Z

dewpoint Dew point temperature

wetbulb Adiabatic saturation temperature

humrat Humidity ratio

relhum Relative humidity

spechum Specific humidity

molarfrac Molar fraction of water vapor

The humidity is specified using two parameters:

How the humidity is specified

The actual value of humidity

The following types are used to characterize the humidity.

WetBulb for wet bulb temperature, actually adiabatic saturation temperature

DewPoint Dew point temperature

RelHum Relative humidity

HumRat Humidity ratio (kg of vapor / kg of dry air)

SpecHum Specific humidity (kg of vapor / kg of moist air)

MolarFrac Molar fraction of water vapor.

Examples

julia> volume(MoistAir, 293.15, WetBulb, 291.15, 101325.0)
0.8464079202783964

julia> volume(MoistAir, 293.15, DewPoint, 291.15, 101325.0)
0.8475219875187474

julia> volume(MoistAir, 293.15, RelHum, 0.7, 101325.0)
0.843889817602806

julia> volume(MoistAir, 20.0u"°C", DewPoint, 60.0u"°F", 1.0u"atm")
0.8449934929585231 kg^-1 m^3

julia> volumem(MoistAir, 293.15, RelHum, 0.5, 93000.0)
0.026199080086890276

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1.0976075893895811 kg m^-3

julia> density(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"lb/inch^3")
3.965358988338535e-5 in^-3 lb

julia> volumem(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"inch^3/kmol")
1.598733210336603e6 in^3 kmol^-1

julia> enthalpy(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
50667.43014746832 kg^-1 J

julia> enthalpym(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
1439.6551689935861 J mol^-1

julia> compressfactor(MoistAir, -90.0u"°C", RelHum, 0.01, 4.5u"MPa")
0.8552758629097985

julia> wetbulb(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
17.0 °C

julia> dewpoint(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa", u"°C")
15.475836053510477 °C

julia> humrat(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.012032930694441925

julia> relhum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.7517801524436909

julia> spechum(MoistAir, 20.0u"°C", WetBulb, 17.0u"°C", 93u"kPa")
0.011889860823189923

Psychro.ww

Psychro.ℓ