Docstrings

Trajectories.@unroll1 — Macro.

@unroll1 for-loop

Unroll the first iteration of a for-loop. Set first to true in the first iteration.

Example:

    @unroll1 for i in 1:10
        if first
            a, state = iterate('A':'Z')
        else
            a, state = iterate('A':'Z', state)
        end
        println(i => a)
    end

Trajectories.APA — Type.

AbstractPairedArray

Simple data structure pairing two arrays, with the first array holding the keys and the second the corresponding values.

For example an array of locations with an array of corresponding measurements or a vector of times with a vector of locations.

Trajectories.AbstractPairedArray — Type.

AbstractPairedArray

Simple data structure pairing two arrays, with the first array holding the keys and the second the corresponding values.

For example an array of locations with an array of corresponding measurements or a vector of times with a vector of locations.

Trajectories.Left

Trajectories.Linear

Trajectories.Right

Trajectories.Trajectories

Trajectories.Trajectory — Type.

Trajectory{S,T} <: AbstractPairedArray

Pairs a linearly ordered (abstract) vector of keys (typically times) with a vector of values (typically locations).

Trajectories._find

Trajectories.checksynchron

Trajectories.eval

Trajectories.include

Trajectories.interpolate — Function.

interpolate(method, X::Trajectory, s)

Interpolate trajectory X using method in [Linear(), Left(), ...] at time s.

Trajectories.issynchron — Function.

issynchron(X, Y) = isequal(keys(X), keys(Y))

Trajectories.mapeach

Trajectories.mapeach!

Trajectories.mapvalues

Trajectories.mapvalues!

Trajectories.piecewise — Function.

piecewise(X::Trajectory, [endtime]) -> tt, xx

If X is a jump process with piecewise constant paths and jumps in X.tt, piecewise returns coordinates path for plotting purposes. The second argument allows to choose the right endtime of the last interval.

Trajectories.trajectory