Docstrings

EchoviewEcs.EchoviewEcs

EchoviewEcs.FLOAT_LABELS — Constant.

FLOAT_LABELS

List of known calibration fields that contain floating point values. Labels NOT in this list are assumed to be Strings.

EchoviewEcs.MYFLOAT

EchoviewEcs.eval

EchoviewEcs.include

EchoviewEcs.load — Function.

load(filename::AbstractString)

Loads an Echoview calibration supplement file, returning a Vector of Dict. Each Dict represents a transducer calibration with keys and values containing the various calibration settings.

See Echoview calibration supplement files.