Docstrings

ClusterTrees.BlockTrees

ClusterTrees.ClusterTrees

ClusterTrees.DepthFirstIterator

ClusterTrees.LevelledTrees

ClusterTrees.Octrees

ClusterTrees.PointerBasedTrees

ClusterTrees.SimpleTrees

ClusterTrees.children — Function.

The expression `children(tree,node)` returns an iterator that will produce
a sequence of node iterators. These values do not have a lot of meaning by
themselves, but can be used in conjunction with the tree object. E.g:

    data(tree, node_itr)
    children(tree, node_itr)

In fact, the node iterators should be regarded as lightweight proxies for
the underlying node and their attached data payload. The node objects
themselves are of limited use for the client programmer as they are an
implementation detail of the specific tree being used.

ClusterTrees.data — Function.

data(tree, node)

Retrieve the data aka payload associated with the given node.

ClusterTrees.depthfirst — Function.

Traverse the tree depth first, executing the function f(tree, node, level) at every node. If f returns false, recursion is halted and the next node on the current level is visited.

ClusterTrees.eval

ClusterTrees.haschildren

ClusterTrees.include

ClusterTrees.insert! — Function.

insert!(tree, parent, data)

Insert a node carrying 'data' as a new child of 'parent'

ClusterTrees.leaves

ClusterTrees.print_tree

ClusterTrees.root — Function.

root(tree)

Return a proxy for the root of the tree.

ClusterTrees.route!

ClusterTrees.update! — Function.

update!(f, tree, node, data, router!)

Algorithm to update or add nodes of the tree. router! and updater! are user supplied functions:

router!(tree, node)

Returns the next candidate node until the node for insertion is reaches. Note that this function potentially created new nodes. Arrival at the destination is indicated by returning the same node that was passed as the second argument.

f(tree, node, data)

Update the destination node node. Typically, data is added in some sense to the data residing at the desitination node.