Docstrings

LocalCoverage.COVDIR — Constant.

Directory for coverage results.

LocalCoverage.LCOVINFO — Constant.

Coverage tracefile.

LocalCoverage.LocalCoverage

LocalCoverage.clean_coverage — Function.

clean_coverage(pkg)

Clean up after generate_coverage.

LocalCoverage.eval

LocalCoverage.generate_coverage — Function.

generate_coverage(pkg; genhtml)

Generate a coverage report for package pkg.

When genhtml, the corresponding external command will be called to generate a HTML report. This can be found in eg the package lcov on Debian/Ubuntu.

*.cov files are near the source files as generated by Julia, everything else is placed in Pkg.dir(pkg, "coverage"). The summary is in Pkg.dir(pkg, "coverage", "lcov.info").

Use clean_coverage for cleaning.

LocalCoverage.include

LocalCoverage.open_coverage — Function.

open_coverage(pkg)

Open the HTML coverage results in a browser for pkg if they exist.

See generate_coverage.