Docstrings

ConvexBodyProximityQueries.ConvexBodyProximityQueries

ConvexBodyProximityQueries.check_degeneracy

ConvexBodyProximityQueries.check_gjk

ConvexBodyProximityQueries.closest_points — Function.

closest_points(p, q, dir; max_iter=100)

Compute the closest points between convex objects p and q if they are not colliding. Provide an initial search direction dir in the space of the problem.

Return result of type Tuple with StaticArrays of the two closest points on each object.

Examples

julia> using StaticArrays
julia> p = SMatrix{2,3}([0.0 0.0 1.0; 0.0 2.0 1.0])
julia> q = SMatrix{2,3}([2.0 2.0 3.0; 0.0 2.0 1.0])
julia> dir = SVector{2}(1.0, 0.0)
julia> closest_points(p, q, dir)
([1.0, 1.0], [2.0, 1.0])

ConvexBodyProximityQueries.collision_detection — Function.

collision_detection(p, q, dir; max_iter=100)

Compute if the convex objects p and q are colliding. Provide an initial search direction dir in the space of the problem.

Examples

julia> using StaticArrays
julia> p = SMatrix{2,3}([0.0 0.0 1.0; 0.0 2.0 1.0])
julia> q = SMatrix{2,3}([2.0 2.0 3.0; 0.0 2.0 1.0])
julia> dir = SVector{2}(1.0, 0.0)
julia> collision_detection(p, q, dir)
false

ConvexBodyProximityQueries.eval

ConvexBodyProximityQueries.findline

ConvexBodyProximityQueries.findsimplex — Function.

findsimplex(simplex)

Compute the new simplices from a pair of given simplices. Return the new search direction. Return a collision flag if the origin was enclosed by the Minkowski simplex.

ConvexBodyProximityQueries.findtetrahedron

ConvexBodyProximityQueries.findtriangle

ConvexBodyProximityQueries.gjk

ConvexBodyProximityQueries.include

ConvexBodyProximityQueries.insertcolumn

ConvexBodyProximityQueries.linecombination

ConvexBodyProximityQueries.minimum_distance — Function.

minimum_distance(p, q, dir; max_iter=100)

Compute the minimum seperating distance between convex objects p and q. Provide an initial search direction dir in the space of the problem.

Examples

julia> using StaticArrays
julia> p = SMatrix{2,3}([0.0 0.0 1.0; 0.0 2.0 1.0])
julia> q = SMatrix{2,3}([2.0 2.0 3.0; 0.0 2.0 1.0])
julia> dir = SVector{2}(1.0, 0.0)
julia> minimum_distance(p, q, dir)
1.0

ConvexBodyProximityQueries.minimum_distance_cond

ConvexBodyProximityQueries.nearestfromsimplex — Function.

nearestfromsimplex(psimplex, qsimplex, dir2origin)

Compute the nearest points between two simplexes given the direction to in origin in the Minkowski difference space

ConvexBodyProximityQueries.pickcolumns

ConvexBodyProximityQueries.proj — Function.

proj(u, v)

Compute the vector projection of vector v onto vector u.

ConvexBodyProximityQueries.reducecolumns

ConvexBodyProximityQueries.support — Function.

support(cvxpoly, dir)

Compute the point of contact between a convex object and its supporting hyperplane defined by the given normal direction as a StaticArray.

ConvexBodyProximityQueries.tolerance_verification — Function.

tolerance_verifcation(p, q, dir, τ; max_iter=100)

Compute if the convex objects p and q are at least τ tolerance apart. Provide an initial search direction dir in the space of the problem.

Examples

julia> using StaticArrays
julia> p = SMatrix{2,3}([0.0 0.0 1.0; 0.0 2.0 1.0])
julia> q = SMatrix{2,3}([2.0 2.0 3.0; 0.0 2.0 1.0])
julia> dir = SVector{2}(1.0, 0.0)
julia> tolerance_verification(p, q, dir, 0.25)
true

ConvexBodyProximityQueries.tolerance_verification_cond

ConvexBodyProximityQueries.trianglecombination