Docstrings

PriorityChannels.PriorityChannel — Type.

PriorityChannel{T}(sz::Int)

Constructs a PriorityChannel with an internal buffer that can hold a maximum of sz objects of type T, each assigned an real-valued priority (low number = higher priority). put! calls on a full channel block until an object is removed with take!.

PriorityChannel(0) constructs an unbuffered channel. put! blocks until a matching take! is called. And vice-versa.

Other constructors:

PriorityChannel(Inf): equivalent to PriorityChannel{Any,Int}(typemax(Int))

PriorityChannel(sz): equivalent to PriorityChannel{Any,Int}(sz)

PriorityChannels.PriorityChannels

PriorityChannels.PriorityElement

PriorityChannels.channeled_tasks — Function.

channeled_tasks(n::Int, funcs...; ctypes=fill(Any,n), csizes=fill(0,n))

A convenience method to create n channels and bind them to tasks started from the provided functions in a single call. Each func must accept n arguments which are the created channels. PriorityChannel types and sizes may be specified via keyword arguments ctypes and csizes respectively. If unspecified, all channels are of type Channel{Any}(0).

Returns a tuple, (Array{Channel}, Array{Task}), of the created channels and tasks.

PriorityChannels.closed_exception

PriorityChannels.eval

PriorityChannels.include

PriorityChannels.isbuffered

PriorityChannels.n_avail

PriorityChannels.ordering