Docstrings

Peaks.Maxima

Peaks.Minima

Peaks.Peaks

Peaks.eval

Peaks.include

Peaks.maxima — Function.

maxima(x[, w=1, includebounds=false])

Find the local maxima of x.

w sets the minimum allowed distance between maxima. If includebounds is true, maxima are allowed to be less than w away from the bounds of x, otherwise, maxima may not be located any closer to the ends of x than w.

Peaks.minima — Function.

minima(x[, w=1, includebounds=false])

Find the local minima of x.

w sets the minimum allowed distance between minima. If includebounds is true, minima are allowed to be less than w away from the bounds of x, otherwise, minima may not be located any closer to the ends of x than w.

Peaks.peakprom — Function.

peakprom(x, ::Maxima[, w=1, minprom=0])

Find all local maxima and maxima prominences in x matching the conditions w and minprom. w sets the minimum allowed distance between maxima. minprom sets the minimum prominence (inclusive) of returned maxima.

Peak prominence is calculated as the distance between the current maxima and the highest of the minimums of the lower and upper bounds. Bounds extend from the next index from the current maxima to the next maxima higher than the current maxima, or the end of the signal, which ever comes first.

Examples

julia> x = rand(1000);

julia> ma, pa = peakprom(x, Maxima());

julia> mi, pi = peakprom(-x, Minima());

julia> @assert (mi == ma) && (pa == pi)

See also: maxima, minima

peakprom(x, ::Minima[, w=1, minprom=0])

Find all local minima and minima prominences in x matching the conditions w and minprom. w sets the minimum allowed distance between minima. minprom sets the minimum prominence (inclusive) of returned minima.

Peak prominence is calculated as the distance between the current minima and the lowest of the maximums of the lower and upper bounds. Bounds extend from the next index from the current minima to the next minima lower than the current minima, or the end of the signal, which ever comes first.

See also: maxima, minima