Docstrings

SchwarzChristoffel.@compat

SchwarzChristoffel.@dotcompat

SchwarzChristoffel.@functorize

SchwarzChristoffel.@recipe

SchwarzChristoffel.@series

SchwarzChristoffel.@shorthands

SchwarzChristoffel.@userplot

SchwarzChristoffel.ABGR

SchwarzChristoffel.ADIN99

SchwarzChristoffel.ADIN99d

SchwarzChristoffel.ADIN99o

SchwarzChristoffel.AGray

SchwarzChristoffel.AGray32

SchwarzChristoffel.AHSI

SchwarzChristoffel.AHSL

SchwarzChristoffel.AHSV

SchwarzChristoffel.ALCHab

SchwarzChristoffel.ALCHuv

SchwarzChristoffel.ALMS

SchwarzChristoffel.ALab

SchwarzChristoffel.ALuv

SchwarzChristoffel.ARGB

SchwarzChristoffel.ARGB32

SchwarzChristoffel.AXYZ

SchwarzChristoffel.AYCbCr

SchwarzChristoffel.AYIQ

SchwarzChristoffel.AbstractBackend

SchwarzChristoffel.AbstractGray

SchwarzChristoffel.AbstractLayout

SchwarzChristoffel.AbstractPlot

SchwarzChristoffel.AbstractRGB

SchwarzChristoffel.AlphaColor

SchwarzChristoffel.AxyY

SchwarzChristoffel.BGR

SchwarzChristoffel.BGRA

SchwarzChristoffel.Color

SchwarzChristoffel.Color3

SchwarzChristoffel.ColorAlpha

SchwarzChristoffel.ColorTypes

SchwarzChristoffel.Colorant

SchwarzChristoffel.ColorantNormed

SchwarzChristoffel.Compat

SchwarzChristoffel.ConformalMap

SchwarzChristoffel.DIN99

SchwarzChristoffel.DIN99A

SchwarzChristoffel.DIN99d

SchwarzChristoffel.DIN99dA

SchwarzChristoffel.DIN99o

SchwarzChristoffel.DIN99oA

SchwarzChristoffel.MapTypes.DerivativeMap — Type.

DerivativeMap(m::ConformalMap)

Constructs new conformal maps from the first and second derivatives of the conformal map m.

These new conformal maps can be evaluated at a single or vector of points just as m is. The first entry in the tuple returned is the first derivative, the second entry is the second derivative.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> dm = DerivativeMap(m);

julia> ζ = [0.1,0.5-0.75im,-0.25-0.3im];

julia> dz, ddz = dm(ζ;inside=true);

julia> dz
3-element Array{Complex{Float64},1}:
  67.2068+76.6284im
 -1.11666+0.544576im
  3.99129-5.30641im

SchwarzChristoffel.Exterior

SchwarzChristoffel.Exterior.ExteriorMap — Type.

ExteriorMap(p::Polygon[;tol::Float64][,ncoeff::Int]) <: ConformalMap

Create a Schwarz-Christoffel map from the interior or exterior of the unit circle to the exterior of polygon p.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p)
Schwarz-Christoffel map of unit circle to exterior of polygon with 4 vertices

ExteriorMap(p;tol=1e-12) manually sets the tolerance to 1e-12 (the default is 1e-8).

ExteriorMap(p;ncoeff=200) manually sets the number of coefficients of negative powers of the multipole expansion of the mapping to 200 (the default is 100).

The resulting map m can be evaluated at a single or vector of points ζ with m(ζ[;inside::Bool]). The points are assumed to lie outside the unit circle, unless the optional argument inside=true, in which case they are assumed to lie inside the circle.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> ζ = [0.1,0.5-0.75im,-0.25-0.3im];

julia> m(ζ;inside=true)
3-element Array{Complex{Float64},1}:
   -6.9344-7.68965im
 0.0439774-1.11249im
   2.41181-0.044779im

julia> ζ = [1.0+3.0im,-2.0-2.0im,0.0+1.1im];

julia> m(ζ)
3-element Array{Complex{Float64},1}:
   0.81614+3.02956im
  -2.25237-2.08523im
 -0.333104+0.975837im

SchwarzChristoffel.Fractional

SchwarzChristoffel.Gray

SchwarzChristoffel.Gray24

SchwarzChristoffel.GrayA

SchwarzChristoffel.HSB

SchwarzChristoffel.HSI

SchwarzChristoffel.HSIA

SchwarzChristoffel.HSL

SchwarzChristoffel.HSLA

SchwarzChristoffel.HSV

SchwarzChristoffel.HSVA

SchwarzChristoffel.MapTypes.InverseMap — Type.

InverseMap(m::ConformalMap)

Constructs the inverse conformal map of the conformal map m.

This inverse conformal map can be evaluated at a single or vector of points. Points should be outside the body. Whether the resulting point in the circle plane is interpreted inside or outside the circle is determined by the optional argument inside, which defaults to false.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> m⁻¹ = InverseMap(m);

julia> ζ = [1.0+3.0im,-2.0-2.0im,0.1+1.1im];

julia> m⁻¹(m(ζ))
3-element Array{Complex{Float64},1}:
  1.0+3.0im
 -2.0-2.0im
  0.1+1.1im

SchwarzChristoffel.Exterior.Jmoment — Function.

Jmoment(m::ConformalMap) -> Float64

Returns the second area moment of the shape described by the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> Jmoment(m)
1.5768333333333333

SchwarzChristoffel.JoukowskiMap

SchwarzChristoffel.Exterior.KarmanTrefftzMap — Type.

KarmanTrefftzMap(ν,ϵ,δ,C[;N = 200]) <: ConformalMap

Create a map from the exterior of the unit circle to the exterior of a Karman-Trefftz airfoil.

The form of the mapping is

\[\frac{z-\nu C}{z+\nu C} = \left(\frac{\tilde{\zeta}-C}{\tilde{\zeta}+C}\right)^\nu\]

where $\tilde{\zeta}$ are the coordinates in an intermediate plane, in which the circle is of radius $a$ and centered at $\epsilon C e^{i\delta}$:

\[\tilde{\zeta} = \epsilon C e^{i\delta} + a \zeta\]

Note that $a/C \geq 1$ and is determined by the choices for $\epsilon$ and $\delta$.

The trailing edge angle, $(2-\nu)\pi$ is specified by $\nu$. The thickness is controlled by $\epsilon C \cos\delta$ and the camber by $\epsilon C \sin\delta$. The airfoil chord length is approximately $4C$. Generally, $\epsilon$ should be much smaller than 1 and $\delta$ between $\pi/2$ and $\pi$.

The resulting map m can be evaluated at a single or a vector of points ζ with m(ζ).

Example

julia> ν = 1.9; ϵ = 0.1; δ = π; C = 0.25;

julia> m = KarmanTrefftzMap(ν,ϵ,δ,C)
Karman-Trefftz map

julia> ζ = [1.0+3.0im,-2.0-2.0im,0.0+1.1im];

julia> m(ζ)
3-element Array{Complex{Float64},1}:
   0.268188+0.764722im
  -0.624265-0.502634im
 -0.0390996+0.126737im

SchwarzChristoffel.LCHab

SchwarzChristoffel.LCHabA

SchwarzChristoffel.LCHuv

SchwarzChristoffel.LCHuvA

SchwarzChristoffel.LMS

SchwarzChristoffel.LMSA

SchwarzChristoffel.Lab

SchwarzChristoffel.LabA

SchwarzChristoffel.Luv

SchwarzChristoffel.LuvA

SchwarzChristoffel.MapTypes

SchwarzChristoffel.Polygons.Polygon — Type.

Polygon(x::Vector{Float64}, y::Vector{Float64})

A polygon defined by its vertices, which must be provided in counter-clockwise order.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0])
Polygon with 4 vertices at
             (-1.0,-1.0) (0.2,-1.0) (1.0,0.5) (-1.0,1.0)
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

SchwarzChristoffel.Polygons

SchwarzChristoffel.Exterior.PowerMap — Type.

PowerMap(c::Vector{ComplexF64}[;N = 200]) <: ConformalMap

Create a power series map from the exterior of the unit circle to the exterior of a shape defined by the power series coefficients c.

The form of the mapping is

\[z(\zeta) = c_{1}\zeta + c_{0} + \sum_{j=1}^{N_{c}} \frac{c_{-j}}{\zeta^{j}}\]

The entries in c correspond as follows: c[1] $\rightarrow c_{1}$, c[2] $\rightarrow c_{0}$, c[3] $\rightarrow c_{-1}$, etc.

The resulting map m can be evaluated at a single or a vector of points ζ with m(ζ).

Example

julia> c = ComplexF64[1,0,1/4];

julia> m = PowerMap(c)
Power series map

julia> ζ = [1.0+3.0im,-2.0-2.0im,0.0+1.1im];

julia> m(ζ)
3-element Array{Complex{Float64},1}:
   1.025+2.925im
 -2.0625-1.9375im
     0.0+0.872727im

SchwarzChristoffel.PowerSeries

SchwarzChristoffel.PowerSeriesDerivatives

SchwarzChristoffel.RGB

SchwarzChristoffel.RGB1

SchwarzChristoffel.RGB24

SchwarzChristoffel.RGB4

SchwarzChristoffel.RGBA

SchwarzChristoffel.RecipeData

SchwarzChristoffel.RecipesBase

SchwarzChristoffel.SchwarzChristoffel

SchwarzChristoffel.Transparent3

SchwarzChristoffel.TransparentColor

SchwarzChristoffel.TransparentGray

SchwarzChristoffel.TransparentRGB

SchwarzChristoffel.XYZ

SchwarzChristoffel.XYZA

SchwarzChristoffel.YCbCr

SchwarzChristoffel.YCbCrA

SchwarzChristoffel.YIQ

SchwarzChristoffel.YIQA

SchwarzChristoffel.Exterior.addedmass — Function.

addedmass(m::ConformalMap) -> Array{Float64,2}

Returns the added mass matrix of the shape described by the conformal mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> addedmass(m)
3×3 Array{Float64,2}:
  0.725129    0.0944902  -1.37387
  0.0944902   3.67634    -0.255119
 -1.37387    -0.255119    3.59231

SchwarzChristoffel.alpha

SchwarzChristoffel.alphacolor

SchwarzChristoffel.Exterior.area — Function.

area(m::ConformalMap) -> Float64

Returns the area of the shape described by the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> area(m)
2.9

julia> c = ComplexF64[1];

julia> m = PowerMap(c);

julia> area(m)
3.141592653589793

SchwarzChristoffel.base_color_type

SchwarzChristoffel.base_colorant_type

SchwarzChristoffel.blue

SchwarzChristoffel.ccolor

SchwarzChristoffel.Exterior.centroid — Function.

centroid(m::ConformalMap) -> ComplexF64

Returns the complex centroid position of the shape described by the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> centroid(m)
-0.20919540229885059 - 0.04022988505747128im

SchwarzChristoffel.Exterior.coefficients — Function.

coefficients(m::ConformalMap) -> Tuple{Vector{ComplexF64},Vector{ComplexF64}}

Returns a tuple of vectors of the complex coefficients of the multipole expansion of the mapping $z(\zeta)$ described by m as well as the coefficients of the square magnitude of the mapping $|z(\zeta)|^2$.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> ccoeff, dcoeff = coefficients(m);

SchwarzChristoffel.color

SchwarzChristoffel.color_type

SchwarzChristoffel.coloralpha

SchwarzChristoffel.comp1

SchwarzChristoffel.comp2

SchwarzChristoffel.comp3

SchwarzChristoffel.eval

SchwarzChristoffel.gamutmax

SchwarzChristoffel.gamutmin

SchwarzChristoffel.gray

SchwarzChristoffel.green

SchwarzChristoffel.include

SchwarzChristoffel.Polygons.interiorangle — Function.

interiorangle(p::Polygon) -> Vector{Float64}

Returns the vector of interior angles (divided by $\pi$) of the polygon p.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> interiorangle(p)
4-element Array{Float64,1}:
 0.5
 0.655958
 0.422021
 0.422021

SchwarzChristoffel.Polygons.isinpoly — Function.

isinpoly(z::Complex128,p::Polygon) -> Bool

Returns true or false depending on whether z is inside or outside polygon p.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> isinpoly(0.0+0.0im,p)
true

julia> isinpoly(1.0+2.0im,p)
false

isinpoly(z::Complex128,p::Polygon,tol::Float64) -> Bool

Returns true if z is inside or within distance tol of polygon p.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> isinpoly(-1.01+0.0im,p)
false

julia> isinpoly(-1.01+0.0im,p,1e-2)
true

SchwarzChristoffel.mapc

SchwarzChristoffel.mapreducec

SchwarzChristoffel.Exterior.moments — Function.

moments(m::ExteriorMap) -> Vector{ComplexF64}

Return the moments of the prevertices for exterior polygon mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> mom = moments(m);

SchwarzChristoffel.myblue

SchwarzChristoffel.mygreen

SchwarzChristoffel.mygreen2

SchwarzChristoffel.Polygons.naca4 — Function.

naca4(cam,pos,t[;np=20][,Zc=0.0+0.0im][,len=1.0]) -> Vector{Complex128}

Generates the vertices of a NACA 4-digit airfoil of chord length 1. The relative camber is specified by cam, the position of maximum camber (as fraction of chord) by pos, and the relative thickness by t.

The optional parameter np specifies the number of points on the upper or lower surface. The optional parameter Zc specifies the mean position of the vertices (which is set to the origin by default). The optional parameter len specifies the chord length.

Example

julia> w = naca4(0.0,0.0,0.12);

julia> p = Polygon(w);

SchwarzChristoffel.Exterior.parameters — Function.

parameters(m::ExteriorMap) -> Tuple{Vector{ComplexF64},ComplexF64}

Returns a tuple of a vector of the prevertices and the complex factor of the exterior polygon mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> prev, C = parameters(m);

julia> prev
4-element Array{Complex{Float64},1}:
       1.0+0.0im
  0.376406-0.926455im
 -0.902383-0.430935im
 -0.186756+0.982406im

SchwarzChristoffel.red

SchwarzChristoffel.reducec

Base.summary — Function.

summary(m::ConformalMap)

Returns a summary of data for a conformal map

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> summary(m)
Schwarz-Christoffel map of unit circle to exterior of polygon with 4 vertices
   vertices: (-1.0,-1.0), (0.2,-1.0), (1.0,0.5), (-1.0,1.0),
   interior angles/π: 0.5, 0.656, 0.422, 0.422,
   prevertices on circle: (1.0,0.0), (0.3764,-0.9265), (-0.9024,-0.4309), (-0.1868,0.9824),
   prevertex angles/π: -0.7291, -0.3519, 0.1291, 0.7111,
   constant = 0.6722 + 0.7669im, accuracy = 1.0e-8,
   number of multipole coefficients = 100

SchwarzChristoffel.Polygons.vertex — Function.

vertex(p::Polygon) -> Vector{Complex128}

Returns the vector of vertices of the polygon p, in complex form.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> vertex(p)
4-element Array{Complex{Float64},1}:
 -1.0-1.0im
  0.2-1.0im
  1.0+0.5im
 -1.0+1.0im

SchwarzChristoffel.xyY

SchwarzChristoffel.xyYA