Docstrings

Zabbix.Zabbix

Zabbix.ZabbixAPI — Type.

Zabbix API

create the zabbix object

julia>z = ZabbixAPI("http://SERVERURL/zabbix/apijsonrpc.php","USERNAME","PASSWORD", true) ZabbixAPI("http://SERVERURL/zabbix/apijsonrpc.php", "USERNAME", "PASSWORD", true, 0, Dict("Content-Type"=>"application/json-rpc"), "2.0")

Note:

Here the verbosity is set to true by default. You may wish to set up by passing false instead of true.

Zabbix.api_version — Function.

Gets the apiinfo.version data from zabbix server

create the zabbix object

julia>z = ZabbixAPI("http://SERVERURL/zabbix/apijsonrpc.php","USERNAME","PASSWORD", true)

get the api version

julia>api_version(z)

INFO: Hitting http://SERVERURL/zabbix/apijsonrpc.php ...

v"3.2.11"

the type returned is actually a VersionNumber

julia>typeof(Zabbix.api_version(zabbix))

VersionNumber

Zabbix.auth_token — Function.

Gets the authentication token from zabbix server

create the zabbix object

julia>z = ZabbixAPI("http://SERVERURL/zabbix/apijsonrpc.php","USERNAME","PASSWORD", true)

get the auth token

julia> auth_token(z)

INFO: Hitting http://SERVERURL/zabbix/apijsonrpc.php ...

"bc86891a1c6c9ef5d41d640eb258a81a"

Zabbix.eval

Zabbix.get_all_hosts — Function.

Gets all host for a user

create the zabbix object

julia>z = ZabbixAPI("http://SERVERURL/zabbix/apijsonrpc.php","USERNAME","PASSWORD", false)

get all hosts for a user

julia>getallhosts(z)

INFO: Getting authentication token ...

INFO: Hitting http://SERVERURL/zabbix/apijsonrpc.php ...

Dict{String,Any} with 3 entries: "id" => 1 "jsonrpc" => "2.0" "result" => Any[Dict{String,Any}(Pair{String,Any}("host", "localhost"),Pair{…

Zabbix.include

Zabbix.make_request — Function.

Make request to the zabbix server

The make_request function requires you to pass methods(aka Zabbix methods like hosts.get etc) and params ie. parameters in a form of a Dict() object. A easy sample is given on Zabbix's official website For references on various methods available in Zabbix, head over to https://www.zabbix.com/documentation/2.2/manual/api/reference

create the zabbix object

julia>z = ZabbixAPI("http://SERVERURL/zabbix/apijsonrpc.php","USERNAME","PASSWORD", true)

create the zabbix method

julia> method = "host.get"

"host.get"

construct the params dict object

julia> params = Dict("output"=>"extend", "filter"=>Dict("host"=>["localhost"]))

Dict{String,Any} with 2 entries: "output" => "extend" "filter" => Dict("host"=>String["localhost"])

create the zabbix object

julia>z = ZabbixAPI("http://SERVERURL/zabbix/apijsonrpc.php","USERNAME","PASSWORD", true)

finally make request

julia> Zabbix.make_request(z, method, params)

INFO: Getting authentication token ...

INFO: Hitting http://SERVERURL/zabbix/apijsonrpc.php ...

INFO: Hitting http://SERVERURL/zabbix/apijsonrpc.php ...

INFO: Updating request id for next API call ...

Dict{String,Any} with 3 entries: "id" => 1 "jsonrpc" => "2.0" "result" => Any[Dict{String,Any}(Pair{String,Any}("lastaccess", "0"),Pair{String,Any}("ipmiprivilege", "2"),Pair{String,Any}("ipmierror…

display all entries for a host

julia> Zabbix.make_request(z, method, params)["result"][1]

INFO: Getting authentication token ...

INFO: Hitting http://SERVERURL/zabbix/apijsonrpc.php ...

INFO: Hitting http://SERVERURL/zabbix/apijsonrpc.php ...

INFO: Updating request id for next API call ...

Dict{String,Any} with 39 entries: "lastaccess" => "0" "ipmiprivilege" => "2" "ipmierrorsfrom" => "0" "snmpavailable" => "0" "templateid" => "0" "disableuntil" => "0" "jmxavailable" => "0" "maintenancefrom" => "0" "tlspskidentity" => "" "available" => "1" "ipmipassword" => "" "tlsaccept" => "1" "name" => "localhost" "tlsissuer" => "" "status" => "0" "maintenancestatus" => "0" "hostid" => "10084" "tlsconnect" => "1" "ipmiavailable" => "0" "description" => "" "errorsfrom" => "0" "maintenancetype" => "0" "error" => "" "ipmiusername" => "" "snmpdisableuntil" => "0" "snmperror" => "" "tlssubject" => "" "maintenanceid" => "0" "host" => "localhost" "jmxerror" => "" "ipmidisableuntil" => "0" "snmperrors_from" => "0" ⋮ => ⋮

get the host id of the zabbix host

julia> Zabbix.make_request(z, method, params)["result"][1]["hostid"]

INFO: Getting authentication token ...

INFO: Hitting http://SERVERURL/zabbix/apijsonrpc.php ...

INFO: Hitting http://SERVERURL/zabbix/apijsonrpc.php ...

INFO: Updating request id for next API call ...

"10084"