Docstrings

StanDump.SPACE

StanDump.Space — Type.

Write a space unless output is requested to be compact.

StanDump.StanDump — Module.

StanDump –- a package for writing data in the CmdStan dump data format.

The single exported function is stan_dump.

StanDump.StanDumpIO — Type.

Wrapper for an IO stream for writing data for use by Stan. See constructor for documentation of slots.

Part of the API, but not exported.

StanDump._dump — Function.

_dump(sd, xs)

Write arguments arguments as data for Stan into sd, passing through strings and characters, and allowing other special objects which are not valid data.

For internal implementation. also defined for objects which are not valid in Stan.

StanDump._dump_vector — Function.

_dump_vector(sd, x)

Dump elements of a vector (or iterable, considered as a vector).

When the element type is <:Integer, dump as Ints, otherwise as Float64; all values are converted for consistency.

StanDump.dump — Function.

dump(sd, xs...)

Write arguments xs... as data for Stan into sd. For internal use.

NOTE: Define methods only for valid Stan objects, using _dump for everything else.

StanDump.eval

StanDump.include

StanDump.is_valid_varname — Function.

is_valid_varname(name)

Test if the argument is valid as a Stan variable name.

NOTE: only basic checks, does not test conflicts with reserved names.

StanDump.stan_dump — Function.

stan_dump(filename, data; force = false, kwargs...)
stan_dump(io, data; kwargs...)
stan_dump(StanDump.StanDumpIO(io; kwargs...), data)

Write data, which is a value that supports pairs (eg a NamedTuple or a Dict) to filename or io.

Using a filename, it will not be overwritten unless force = true is specified.

Keyword arguments are passed to StanDumpIO to govern formatting (most users should not care about this, except for debugging purposes).

Example

julia> stan_dump(stdout, (N = 1, a = 1:5, b = ones(2, 2)))
N <- 1
a <- 1:5
b <- structure(c(1.0, 1.0, 1.0, 1.0), .Dim = c(2, 2))