Docstrings

Coulter.-

Coulter.Coulter

Coulter.CoulterCounterRun — Type.

A simplified representation of a coulter counter run

Coulter._find_peaks — Function.

_find_peaks(xs, ys; minx, miny)

Finds prominent peaks in ys and returns the values from xs corresponding to the location of these peaks. minx and miny can be used to exclude peaks that are too small in x or y.

Coulter.diameter — Function.

diameter(volume)

Given the volume of a sphere, compute its diameter

Coulter.eval

Coulter.extract_peak

Coulter.extract_peak!

Coulter.extract_peak_interval — Function.

extract_peak_interval(run; α, n)

Takes a CoulterCounterRun run and fits a kernel density estimate to smooth out binning effects by the machine and then finds the largest peak. It returns the diameter corresponding to the location of this peak as well as the max and min diameters of the confidence interval defined by α. This interval is generated on the empirically by bootstrapping the data n times.

Coulter.include

Coulter.loadZ2 — Function.

loadZ2(filename::String, sample::String)

Loads filename and assigns it a sample, returns a CoulterCounterRun object

Coulter.load_folder — Function.

load_folder(folder)

Loads all coulter runs in a folder into a dictionary where the keys are the first parts of the filenames separated by an underscore.

Coulter.repvec — Function.

repvec(orig, reps)

Repeats the items in the first vector by the corresponding number of times in the second vector. Essentially the inverse operation of hist

Coulter.theme

Coulter.volume — Function.

volume(diameter)

Given the diameter of a sphere, return its volume