Docstrings

MMTF.BASE_URL

MMTF.CHAIN_LEN

MMTF.COORD_DIVIDER

MMTF.MAX_SHORT

MMTF.MIN_SHORT

MMTF.MMTF

MMTF.NULL_BYTE

MMTF.OCC_B_FACTOR_DIVIDER

MMTF.addheader — Function.

Add the header to the appropriate array

MMTF.bytes — Function.

Convert an integer array into a byte arrays.

MMTF.chars — Function.

Convert integers to chars.

MMTF.decodearray — Function.

Parse the header of an input byte array and then decode using the input array, the codec and the appropirate parameter. input_array: the array to be decoded

MMTF.decodechainlist — Function.

Convert a list of bytes to a list of strings. Each string is of length mmtf.CHAIN_LEN

MMTF.decodedata — Function.

Decodes the MsgPack unpacked dict of the MMTF file.

MMTF.deltadecode — Function.

A function to delta decode an int array. in_array: the input array of integers

MMTF.deltaencode — Function.

A function to delta decode an int array. in_array: the input array to be delta encoded

MMTF.deltarecursivefloatdecode

MMTF.deltarecursivefloatencode

MMTF.encodearray — Function.

Encode the array using the method and then add the header to this array. input_array: the array to be encoded codec: the integer index of the codec to use param: the integer parameter to use in the function

MMTF.encodechainlist — Function.

Convert a list of strings to a list of byte arrays.

MMTF.encodedata — Function.

Encodes the MMTF Dict.

MMTF.eval

MMTF.fetchmmtf — Function.

Fetches a file from RCSB server and decodes it into a MMTFDict. pdbid: ID of the PDB file to be fetched and decoded.

MMTF.floats — Function.

Convert integers to floats by division. in_ints: the integer array divider: the divider

MMTF.fourbyteintdecode

MMTF.fourbyteintencode

MMTF.htonofarray

MMTF.include

MMTF.intdecode

MMTF.intencode

MMTF.ints — Function.

Convert a byte array into an integer array. The number of bytes forming an integer is defined by num in_bytes: the input bytes num: the number of bytes per int

Convert floating points to integers using a multiplier. in_floats: the input floats multiplier: the multiplier to be used for conversion. Corresponds to the precisison.

Convert an array of chars to an array of ints.

MMTF.ntohofarray

MMTF.parseheader — Function.

Parse the header and return it along with the input array minus the header. Returns the codec, the length of the decoded array, the parameter and the remainder of the array

MMTF.parsemmtf — Function.

Parses the MMTF file into a MMTFDict. filepath: Path of the input file.

MMTF.recursiveindexdecode — Function.

Unpack an array of integers using recursive indexing.

MMTF.recursiveindexencode — Function.

Pack an integer array using recursive indexing.

MMTF.runlengthchardecode

MMTF.runlengthcharencode

MMTF.runlengthdecode — Function.

A function to run length decode an int array. in_array: the input array of integers

MMTF.runlengthdeltaintdecode

MMTF.runlengthdeltaintencode

MMTF.runlengthencode — Function.

A function to run length decode an int array. in_array: the inptut array of integers

MMTF.runlengthfloatdecode

MMTF.runlengthfloatencode

MMTF.stringdecode

MMTF.stringencode

MMTF.writemmtf — Function.

Writes the MMTFDict back to a file. MMTFDict: Dict containing the decoded MMTF data to be written to file. filepath: Path to write the file