Docstrings

BayesianNonparametrics.AbstractHyperparam — Type.

abstract Hyperparameters

BayesianNonparametrics.AbstractModelBuffer — Type.

Abstract Model Buffer Object

BayesianNonparametrics.AbstractModelData — Type.

Abstract Model Data Object

BayesianNonparametrics.BayesianNonparametrics

BayesianNonparametrics.BetaBernoulli

BayesianNonparametrics.ConjugatePostDistribution

BayesianNonparametrics.ContinuousMultivariateConjugatePostDistribution

BayesianNonparametrics.ContinuousUnivariateConjugatePostDistribution

BayesianNonparametrics.DPM

BayesianNonparametrics.DPMBuffer

BayesianNonparametrics.DPMData — Type.

Dirichlet Process Mixture Model Data Object

BayesianNonparametrics.DPMHyperparam — Type.

Dirichlet Process Mixture Model Hyperparameters

BayesianNonparametrics.DirichletCategorical

BayesianNonparametrics.DirichletMultinomial

BayesianNonparametrics.DiscreteMultivariateConjugatePostDistribution

BayesianNonparametrics.DiscreteUnivariateConjugatePostDistribution

BayesianNonparametrics.GammaNormal — Type.

GammaNormal(; μ0 = 0.0, λ0 = 1.0, α0 = 1.0, β0 = 1.0)

Normal-Gamma distribution

A Normal-Gamma distribution is the conjugate prior of a Normal distribution with unknown mean and precision.

Paramters

μ0: location

λ0 > 0: number of pseudo-observations

α0 > 0

β0 > 0

Example:

d = GammaNormal()

BayesianNonparametrics.GaussianDiagonal

BayesianNonparametrics.Gibbs

BayesianNonparametrics.HDP

BayesianNonparametrics.HDPBuffer

BayesianNonparametrics.HDPData — Type.

Hierarchical Dirichlet Process Mixture Model Data Object

BayesianNonparametrics.HDPHyperparam — Type.

Hierarchical Dirichlet Process Mixture Model Hyperparameters

BayesianNonparametrics.IncrementalInitialisation

BayesianNonparametrics.InitialisationType

BayesianNonparametrics.KMeansInitialisation

BayesianNonparametrics.ModelType

BayesianNonparametrics.MultivariateConjugatePostDistribution

BayesianNonparametrics.NormalNormal

BayesianNonparametrics.PosteriorInference

BayesianNonparametrics.PrecomputedInitialisation

BayesianNonparametrics.RandomInitialisation

BayesianNonparametrics.SliceSampler

BayesianNonparametrics.UnivariateConjugatePostDistribution

BayesianNonparametrics.VCM

BayesianNonparametrics.WishartGaussian — Type.

WishartGaussian(μ0, κ0, ν0, Σ0)

Gaussian-inverse-Wishart distribution

A Gaussian-inverse-Wishart distribution is the conjugate prior of a multivariate normal distribution with unknown mean and covariance matrix.

Parameters

μ0, Dx1: location

κ0 > 0: number of pseudo-observations

ν0 > D-1: degrees of freedom

Σ0 > 0, DxD: scale matrix

Example

julia> (N, D) = size(X)
julia> μ0 = mean(X, dims = 1)
julia> d = WishartGaussian(μ0, 1.0, 2*D, cov(x)) 

BayesianNonparametrics.add — Function.

add(d::ConjugatePostDistribution, X)

Add data to Posterior Distribution.

BayesianNonparametrics.add! — Function.

add!(d::ConjugatePostDistribution, X)

Add data to Posterior Distribution (inplace).

BayesianNonparametrics.bars

BayesianNonparametrics.bloobs

BayesianNonparametrics.eval

BayesianNonparametrics.extractpointestimate

BayesianNonparametrics.gibbs! — Function.

Single iteration of collabsed Gibbs sampling using CRP.

Single Gibbs sweep of HDP training using beta variables.

BayesianNonparametrics.include

BayesianNonparametrics.init — Function.

HDP initialization using random assignments.

BayesianNonparametrics.isdistempty — Function.

isdistempty(d::ConjugatePostDistribution)

Check if distribution is empty (contains no samples).

BayesianNonparametrics.logpostpred — Function.

logpostpred(d::ConjugatePostDistribution, X)

Compute log posterior predictive.

Log PMF for MultinomialDirichlet.

Log PMF for MultinomialDirichlet.

Log PMF for BernoulliBeta.

BayesianNonparametrics.point_estimate_hclust — Function.

point_estimate_avg(psm::Array{Float64, 2})

Find optimal partition which minimizes the lower bound to the Variation of Information obtain from Jensen's inequality where the expectation and log are reversed.

Code based on R implementation by Sara Wade <sara.wade@eng.cam.ac.uk>

BayesianNonparametrics.pointestimate — Function.

point_estimate(psm::Array{Float64, 2})

Find optimal partition which minimizes the lower bound to the Variation of Information obtain from Jensen's inequality where the expectation and log are reversed.

Code based on R implementation by Sara Wade <sara.wade@eng.cam.ac.uk>

BayesianNonparametrics.posteriorParameters — Function.

posteriorParameters(d::ConjugatePostDistribution)

Compute posterior distribution parameters.

BayesianNonparametrics.randnumtable

BayesianNonparametrics.randomindex — Function.

randomindex(p::Vector{Float64})

Randomly select an index propotional to its probability.

BayesianNonparametrics.remove — Function.

remove(d::ConjugatePostDistribution, X::AbstractArray)

Remove data from Posterior Distribution.

BayesianNonparametrics.remove! — Function.

remove!(d::ConjugatePostDistribution, X)

Remove data from Posterior Distribution (inplace).

BayesianNonparametrics.resampleα — Function.

resampleα(α::Float64, N::Int, K::Int)

Resample α given the approach by Escobar and West. (page 585)

BayesianNonparametrics.sampleparameters!

BayesianNonparametrics.slicemap

BayesianNonparametrics.slicesampling!

BayesianNonparametrics.tlogpdf — Function.

Log PDF of Generalized student-t Distribution.

BayesianNonparametrics.train

BayesianNonparametrics.updateenergy! — Function.

Compute Energy of model for given data

BayesianNonparametrics.variation_of_information_lb — Function.

variation_of_information_lb()

Computes the lower bound to the posterior expected Variation of Information.

variation_of_information_lb()

Computes the lower bound to the posterior expected Variation of Information.

BayesianNonparametrics.variationalbayes!