Docstrings

GoogleCloudObjectStores.Bucket

GoogleCloudObjectStores.GoogleCloudObjectStore

GoogleCloudObjectStores.GoogleCloudObjectStores

GoogleCloudObjectStores.GoogleObjectStore

GoogleCloudObjectStores.Object

GoogleCloudObjectStores.ObjectStore

GoogleCloudObjectStores.ObjectStores

GoogleCloudObjectStores.Permission

GoogleCloudObjectStores._create! — Function.

Create bucket. If successful return nothing, else return an error message as a String.

Create object. If successful return nothing, else return an error message as a String.

v is either:

NamedTuple: (mimetype="application/json", value="some value")

An arbitrary value

GoogleCloudObjectStores._delete! — Function.

Delete bucket. If successful return nothing, else return an error message as a String.

Delete object. If successful return nothing, else return an error message as a String.

GoogleCloudObjectStores._isbucket

GoogleCloudObjectStores._isobject

GoogleCloudObjectStores._read — Function.

List the contents of a specific bucket. If successful return (true, value), else return (false, errormessage::String).

Read object. If successful return (true, value), else return (false, errormessage::String).

GoogleCloudObjectStores.createbucket!

GoogleCloudObjectStores.delete!

GoogleCloudObjectStores.deletebucket!

GoogleCloudObjectStores.eval

GoogleCloudObjectStores.getindex

GoogleCloudObjectStores.getpermission

GoogleCloudObjectStores.haspermission

GoogleCloudObjectStores.include

GoogleCloudObjectStores.isbucket

GoogleCloudObjectStores.isobject

GoogleCloudObjectStores.listcontents

GoogleCloudObjectStores.permissions_conflict

GoogleCloudObjectStores.setexpiry!

GoogleCloudObjectStores.setindex!

GoogleCloudObjectStores.setpermission!