Docstrings

IndentWrappers.IndentWrapper

IndentWrappers.IndentWrappers — Module.

Wrapper type for indentation management for plain text printing.

The single exported function is indent, see its docstring for usage.

IndentWrappers._write_spaces

IndentWrappers.eval

IndentWrappers.include

IndentWrappers.indent — Function.

indent(io, spaces)

Return a wrapper around io that prepends each written to the stream with the given number of spaces.

It is recommended that indent is chained together in a functional manner. Blocks should always begin with a newline and end without one.

Example

julia> let io = stdout
           print(io, "toplevel")
           let io = indent(io, 4)
               print(io, '
', "- level1")
               let io = indent(io, 4)
                   print(io, '
', "- level 2")
               end
           end
       end
toplevel
    - level1
        - level 2