Docstrings

Qutilities.LOG

Qutilities.Qutilities

Qutilities.S_renyi — Function.

S_renyi(rho::AbstractMatrix, alpha::Real=2)

Order alpha Rényi entropy of rho.

The alpha parameter may have any value on the interval [0, Inf] except 1. It defaults to 2.

Qutilities.S_vn — Function.

S_vn(rho::AbstractMatrix)

Von Neumann entropy of rho.

Qutilities.binent — Function.

binent(x::Real)

Binary entropy of x.

Qutilities.concurrence — Function.

concurrence(rho::AbstractMatrix)

Concurrence of a mixed state rho for two qubits in the standard basis.

Ref: Wootters, W. K. (1998). Entanglement of formation of an arbitrary state of two qubits. Physical Review Letters, 80(10), 2245.

Qutilities.concurrence_lb — Function.

concurrence_lb(rho::AbstractMatrix)

Lower bound on the concurrence for two qubits in the mixed state rho in the standard basis.

Ref: Mintert, F., & Buchleitner, A. (2007). Observable entanglement measure for mixed quantum states. Physical Review Letters, 98(14), 140505.

Qutilities.eval

Qutilities.formation — Function.

formation(C::Real)

Entanglement of formation for two qubits with concurrence C.

Ref: Wootters, W. K. (1998). Entanglement of formation of an arbitrary state of two qubits. Physical Review Letters, 80(10), 2245.

formation(rho::AbstractMatrix)

Entanglement of formation for two qubits in the mixed state rho in the standard basis.

Ref: Wootters, W. K. (1998). Entanglement of formation of an arbitrary state of two qubits. Physical Review Letters, 80(10), 2245.

Qutilities.hermitize — Function.

hermitize(rho::AbstractMatrix)

Make rho Hermitian, but carefully.

Qutilities.include

Qutilities.mutinf — Function.

mutinf(rho::AbstractMatrix, S::Function=S_vn)

Mutual information of rho using the entropy function S.

Qutilities.negativity — Function.

negativity(rho::AbstractMatrix)

Logarithmic negativity for a symmetric bipartition of rho.

Ref: Plenio, M. B. (2005). Logarithmic negativity: A full entanglement monotone that is not convex. Physical Review Letters, 95(9), 090503.

Qutilities.nonneg — Function.

nonneg(x::Real)

x if x is non-negative; zero if x is negative.

Qutilities.ptrace — Function.

ptrace(rho::AbstractMatrix{T}, dims, which::Int)

Partial trace of rho along dimension which from dims.

ptrace(rho::AbstractMatrix, which::Int=2)

Partial trace of rho along dimension which.

rho is split into halves and the trace is over the second half by default.

Qutilities.ptranspose — Function.

ptranspose(rho::AbstractMatrix, dims, which::Int)

Partial transpose of rho along dimension which from dims.

ptranspose(rho::AbstractMatrix, which::Int=2)

Partial transpose of rho along dimension which.

rho is split into halves and the transpose is over the second half by default.

Qutilities.purity — Function.

purity(rho::AbstractMatrix)

Purity of rho.

Qutilities.shannon — Function.

shannon(xs::AbstractVector)

Shannon entropy of xs.

Qutilities.sigma_x

Qutilities.sigma_y

Qutilities.sigma_z

Qutilities.spinflip — Function.

spinflip(rho::AbstractMatrix)

Wootters spin-flip operation for two qubits in the mixed state rho in the standard basis.

Ref: Wootters, W. K. (1998). Entanglement of formation of an arbitrary state of two qubits. Physical Review Letters, 80(10), 2245.