Docstrings

DynamicHMCModels.+

DynamicHMCModels.-

DynamicHMCModels./

DynamicHMCModels.@animate

DynamicHMCModels.@colorant_str

DynamicHMCModels.@csv2_str

DynamicHMCModels.@csv_str

DynamicHMCModels.@df

DynamicHMCModels.@gif

DynamicHMCModels.@layout

DynamicHMCModels.@pack

DynamicHMCModels.@pack!

DynamicHMCModels.@recipe

DynamicHMCModels.@series

DynamicHMCModels.@shorthands

DynamicHMCModels.@tsv_str

DynamicHMCModels.@unpack

DynamicHMCModels.@userplot

DynamicHMCModels.@with_kw

DynamicHMCModels.@with_kw_noshow

DynamicHMCModels.@wsv_str

DynamicHMCModels.ABGR

DynamicHMCModels.ADIN99

DynamicHMCModels.ADIN99d

DynamicHMCModels.ADIN99o

DynamicHMCModels.ADgradient

DynamicHMCModels.AGray

DynamicHMCModels.AGray32

DynamicHMCModels.AHSI

DynamicHMCModels.AHSL

DynamicHMCModels.AHSV

DynamicHMCModels.ALCHab

DynamicHMCModels.ALCHuv

DynamicHMCModels.ALMS

DynamicHMCModels.ALab

DynamicHMCModels.ALuv

DynamicHMCModels.ARGB

DynamicHMCModels.ARGB32

DynamicHMCModels.ARPACKException

DynamicHMCModels.AXYZ

DynamicHMCModels.AYCbCr

DynamicHMCModels.AYIQ

DynamicHMCModels.AbstractBackend

DynamicHMCModels.AbstractCategoricalArray

DynamicHMCModels.AbstractCategoricalMatrix

DynamicHMCModels.AbstractCategoricalVector

DynamicHMCModels.AbstractDataFrame

DynamicHMCModels.AbstractGray

DynamicHMCModels.AbstractHistogram

DynamicHMCModels.AbstractLayout

DynamicHMCModels.AbstractMissingCategoricalArray

DynamicHMCModels.AbstractMissingCategoricalMatrix

DynamicHMCModels.AbstractMissingCategoricalVector

DynamicHMCModels.AbstractMixtureModel

DynamicHMCModels.AbstractMvNormal

DynamicHMCModels.AbstractPlot

DynamicHMCModels.AbstractRGB

DynamicHMCModels.AbstractSerializer

DynamicHMCModels.AbstractWeights

DynamicHMCModels.Adjoint

DynamicHMCModels.AlphaColor

DynamicHMCModels.AnalyticWeights

DynamicHMCModels.Animation

DynamicHMCModels.Arcsine

DynamicHMCModels.AxyY

DynamicHMCModels.BGR

DynamicHMCModels.BGRA

DynamicHMCModels.BLAS

DynamicHMCModels.Bernoulli

DynamicHMCModels.Beta

DynamicHMCModels.BetaBinomial

DynamicHMCModels.BetaPrime

DynamicHMCModels.BezierCurve

DynamicHMCModels.Bidiagonal

DynamicHMCModels.Binomial

DynamicHMCModels.Biweight

DynamicHMCModels.BunchKaufman

DynamicHMCModels.CIE1931JV_CMF

DynamicHMCModels.CIE1931J_CMF

DynamicHMCModels.CIE1931_CMF

DynamicHMCModels.CIE1964_CMF

DynamicHMCModels.CSV

DynamicHMCModels.Categorical

DynamicHMCModels.CategoricalArray

DynamicHMCModels.CategoricalArrays

DynamicHMCModels.CategoricalMatrix

DynamicHMCModels.CategoricalPool

DynamicHMCModels.CategoricalString

DynamicHMCModels.CategoricalValue

DynamicHMCModels.CategoricalVector

DynamicHMCModels.Cauchy

DynamicHMCModels.Chains

DynamicHMCModels.Chi

DynamicHMCModels.Chisq

DynamicHMCModels.Cholesky

DynamicHMCModels.CholeskyPivoted

DynamicHMCModels.CoefTable

DynamicHMCModels.Color

DynamicHMCModels.Color3

DynamicHMCModels.ColorAlpha

DynamicHMCModels.ColorGradient

DynamicHMCModels.ColorTypes

DynamicHMCModels.Colorant

DynamicHMCModels.ColorantNormed

DynamicHMCModels.Colors

DynamicHMCModels.Continuous

DynamicHMCModels.ContinuousDistribution

DynamicHMCModels.ContinuousMatrixDistribution

DynamicHMCModels.ContinuousMultivariateDistribution

DynamicHMCModels.ContinuousUnivariateDistribution

DynamicHMCModels.ConvergenceException

DynamicHMCModels.CorrCholeskyFactor

DynamicHMCModels.Cosine

DynamicHMCModels.CustomTransform

DynamicHMCModels.DE_2000

DynamicHMCModels.DE_94

DynamicHMCModels.DE_AB

DynamicHMCModels.DE_BFD

DynamicHMCModels.DE_CMC

DynamicHMCModels.DE_DIN99

DynamicHMCModels.DE_DIN99d

DynamicHMCModels.DE_DIN99o

DynamicHMCModels.DE_JPC79

DynamicHMCModels.DIN99

DynamicHMCModels.DIN99A

DynamicHMCModels.DIN99d

DynamicHMCModels.DIN99dA

DynamicHMCModels.DIN99o

DynamicHMCModels.DIN99oA

DynamicHMCModels.DataFrame

DynamicHMCModels.DataFrameRow

DynamicHMCModels.DataFrames

DynamicHMCModels.DelimitedFiles

DynamicHMCModels.DiagNormal

DynamicHMCModels.DiagNormalCanon

DynamicHMCModels.Diagonal

DynamicHMCModels.DiffResults

DynamicHMCModels.Dirichlet

DynamicHMCModels.DirichletMultinomial

DynamicHMCModels.Discrete

DynamicHMCModels.DiscreteDistribution

DynamicHMCModels.DiscreteMatrixDistribution

DynamicHMCModels.DiscreteMultivariateDistribution

DynamicHMCModels.DiscreteUniform

DynamicHMCModels.DiscreteUnivariateDistribution

DynamicHMCModels.Distribution

DynamicHMCModels.Distributions

DynamicHMCModels.DoubleExponential

DynamicHMCModels.DynamicHMC

DynamicHMCModels.DynamicHMCModels

DynamicHMCModels.EBFMI

DynamicHMCModels.ECDF

DynamicHMCModels.EdgeworthMean

DynamicHMCModels.EdgeworthSum

DynamicHMCModels.EdgeworthZ

DynamicHMCModels.Eigen

DynamicHMCModels.EmpiricalUnivariateDistribution

DynamicHMCModels.Epanechnikov

DynamicHMCModels.Erlang

DynamicHMCModels.Estimator

DynamicHMCModels.EuclideanKE

DynamicHMCModels.Exponential

DynamicHMCModels.FDist

DynamicHMCModels.Factorization

DynamicHMCModels.FisherNoncentralHypergeometric

DynamicHMCModels.Formatted

DynamicHMCModels.ForwardDiff

DynamicHMCModels.Fractional

DynamicHMCModels.Frechet

DynamicHMCModels.FrequencyWeights

DynamicHMCModels.FullNormal

DynamicHMCModels.FullNormalCanon

DynamicHMCModels.GR

DynamicHMCModels.Gamma

DynamicHMCModels.GaussianKE

DynamicHMCModels.GeneralizedEigen

DynamicHMCModels.GeneralizedExtremeValue

DynamicHMCModels.GeneralizedPareto

DynamicHMCModels.GeneralizedSVD

DynamicHMCModels.GeneralizedSchur

DynamicHMCModels.Geometric

DynamicHMCModels.Gray

DynamicHMCModels.Gray24

DynamicHMCModels.GrayA

DynamicHMCModels.GroupApplied

DynamicHMCModels.GroupedDataFrame

DynamicHMCModels.Gumbel

DynamicHMCModels.HSB

DynamicHMCModels.HSI

DynamicHMCModels.HSIA

DynamicHMCModels.HSL

DynamicHMCModels.HSLA

DynamicHMCModels.HSV

DynamicHMCModels.HSVA

DynamicHMCModels.Hermitian

DynamicHMCModels.Hessenberg

DynamicHMCModels.Histogram

DynamicHMCModels.Hypergeometric

DynamicHMCModels.I

DynamicHMCModels.InverseGamma

DynamicHMCModels.InverseGaussian

DynamicHMCModels.InverseWishart

DynamicHMCModels.IsoNormal

DynamicHMCModels.IsoNormalCanon

DynamicHMCModels.KSDist

DynamicHMCModels.KSOneSided

DynamicHMCModels.KW

DynamicHMCModels.KineticEnergy

DynamicHMCModels.Kolmogorov

DynamicHMCModels.L1dist

DynamicHMCModels.L2dist

DynamicHMCModels.LAPACK

DynamicHMCModels.LAPACKException

DynamicHMCModels.LCHab

DynamicHMCModels.LCHabA

DynamicHMCModels.LCHuv

DynamicHMCModels.LCHuvA

DynamicHMCModels.LDLt

DynamicHMCModels.LMS

DynamicHMCModels.LMSA

DynamicHMCModels.LQ

DynamicHMCModels.LU

DynamicHMCModels.Lab

DynamicHMCModels.LabA

DynamicHMCModels.Laplace

DynamicHMCModels.LevelsException

DynamicHMCModels.Levy

DynamicHMCModels.LinearAlgebra

DynamicHMCModels.Linfdist

DynamicHMCModels.LocationScale

DynamicHMCModels.LogDensityProblems

DynamicHMCModels.LogDensityRejectErrors

DynamicHMCModels.LogNormal

DynamicHMCModels.Logistic

DynamicHMCModels.LowerTriangular

DynamicHMCModels.Luv

DynamicHMCModels.LuvA

DynamicHMCModels.MCMCChain

DynamicHMCModels.MCMCDiagnostics

DynamicHMCModels.MLEstimator

DynamicHMCModels.MSC

DynamicHMCModels.MatrixDistribution

DynamicHMCModels.Matrixvariate

DynamicHMCModels.Missing

DynamicHMCModels.MissingCategoricalArray

DynamicHMCModels.MissingCategoricalMatrix

DynamicHMCModels.MissingCategoricalVector

DynamicHMCModels.MissingException

DynamicHMCModels.Missings

DynamicHMCModels.MixtureModel

DynamicHMCModels.Multinomial

DynamicHMCModels.Multivariate

DynamicHMCModels.MultivariateDistribution

DynamicHMCModels.MultivariateMixture

DynamicHMCModels.MultivariateNormal

DynamicHMCModels.MvLogNormal

DynamicHMCModels.MvNormal

DynamicHMCModels.MvNormalCanon

DynamicHMCModels.MvNormalKnownCov

DynamicHMCModels.MvTDist

DynamicHMCModels.NUTS

DynamicHMCModels.NUTS_Transition

DynamicHMCModels.NUTS_init_tune_mcmc

DynamicHMCModels.NUTS_statistics

DynamicHMCModels.NegativeBinomial

DynamicHMCModels.NonMatrixDistribution

DynamicHMCModels.NoncentralBeta

DynamicHMCModels.NoncentralChisq

DynamicHMCModels.NoncentralF

DynamicHMCModels.NoncentralHypergeometric

DynamicHMCModels.NoncentralT

DynamicHMCModels.Normal

DynamicHMCModels.NormalCanon

DynamicHMCModels.NormalInverseGaussian

DynamicHMCModels.OHLC

DynamicHMCModels.OrderedLevelsException

DynamicHMCModels.Parameters

DynamicHMCModels.Pareto

DynamicHMCModels.PlotThemes

DynamicHMCModels.PlotUtils

DynamicHMCModels.Plots

DynamicHMCModels.Poisson

DynamicHMCModels.PoissonBinomial

DynamicHMCModels.PosDefException

DynamicHMCModels.ProbabilityWeights

DynamicHMCModels.QQPair

DynamicHMCModels.QR

DynamicHMCModels.QRPivoted

DynamicHMCModels.RDatasets

DynamicHMCModels.RGB

DynamicHMCModels.RGB1

DynamicHMCModels.RGB24

DynamicHMCModels.RGB4

DynamicHMCModels.RGBA

DynamicHMCModels.RankDeficientException

DynamicHMCModels.Rayleigh

DynamicHMCModels.RealInterval

DynamicHMCModels.RecipeData

DynamicHMCModels.RecipesBase

DynamicHMCModels.RegressionModel

DynamicHMCModels.ReportIO

DynamicHMCModels.ReportSilent

DynamicHMCModels.SVD

DynamicHMCModels.Sampleable

DynamicHMCModels.Schur

DynamicHMCModels.ScriptEntry

DynamicHMCModels.Segments

DynamicHMCModels.Semicircle

DynamicHMCModels.Serialization

DynamicHMCModels.Serializer

DynamicHMCModels.Shape

DynamicHMCModels.SingularException

DynamicHMCModels.Skellam

DynamicHMCModels.StatisticalModel

DynamicHMCModels.StatisticalRethinking

DynamicHMCModels.StatsBase

DynamicHMCModels.StatsFuns

DynamicHMCModels.StatsPlots

DynamicHMCModels.SubDataFrame

DynamicHMCModels.SufficientStats

DynamicHMCModels.Surface

DynamicHMCModels.SymTriangularDist

DynamicHMCModels.SymTridiagonal

DynamicHMCModels.Symmetric

DynamicHMCModels.TDist

DynamicHMCModels.TransformVariables

DynamicHMCModels.TransformedLogDensity

DynamicHMCModels.Transparent3

DynamicHMCModels.TransparentColor

DynamicHMCModels.TransparentGray

DynamicHMCModels.TransparentRGB

DynamicHMCModels.Transpose

DynamicHMCModels.TriangularDist

DynamicHMCModels.Tridiagonal

DynamicHMCModels.Triweight

DynamicHMCModels.Truncated

DynamicHMCModels.TruncatedNormal

DynamicHMCModels.Uniform

DynamicHMCModels.UniformScaling

DynamicHMCModels.UnitLowerTriangular

DynamicHMCModels.UnitUpperTriangular

DynamicHMCModels.UnitVector

DynamicHMCModels.Univariate

DynamicHMCModels.UnivariateDistribution

DynamicHMCModels.UnivariateGMM

DynamicHMCModels.UnivariateMixture

DynamicHMCModels.UpperTriangular

DynamicHMCModels.ValueSupport

DynamicHMCModels.VariateForm

DynamicHMCModels.VonMises

DynamicHMCModels.VonMisesFisher

DynamicHMCModels.WalleniusNoncentralHypergeometric

DynamicHMCModels.Weibull

DynamicHMCModels.Weights

DynamicHMCModels.Wishart

DynamicHMCModels.XYZ

DynamicHMCModels.XYZA

DynamicHMCModels.YCbCr

DynamicHMCModels.YCbCrA

DynamicHMCModels.YIQ

DynamicHMCModels.YIQA

DynamicHMCModels.ZeroMeanDiagNormal

DynamicHMCModels.ZeroMeanDiagNormalCanon

DynamicHMCModels.ZeroMeanFullNormal

DynamicHMCModels.ZeroMeanFullNormalCanon

DynamicHMCModels.ZeroMeanIsoNormal

DynamicHMCModels.ZeroMeanIsoNormalCanon

DynamicHMCModels.\

DynamicHMCModels.adapted_grid

DynamicHMCModels.add_backend

DynamicHMCModels.add_theme

DynamicHMCModels.addcounts!

DynamicHMCModels.adjoint

DynamicHMCModels.adjoint!

DynamicHMCModels.adjr2

DynamicHMCModels.adjr²

DynamicHMCModels.aggregate

DynamicHMCModels.aic

DynamicHMCModels.aicc

DynamicHMCModels.aliases

DynamicHMCModels.allowmissing

DynamicHMCModels.allowmissing!

DynamicHMCModels.alpha

DynamicHMCModels.alphacolor

DynamicHMCModels.andrewsplot

DynamicHMCModels.andrewsplot!

DynamicHMCModels.animate

DynamicHMCModels.annotate!

DynamicHMCModels.arrow

DynamicHMCModels.as

DynamicHMCModels.asℝ

DynamicHMCModels.asℝ₊

DynamicHMCModels.asℝ₋

DynamicHMCModels.as𝕀

DynamicHMCModels.attr!

DynamicHMCModels.autcorplot

DynamicHMCModels.autocor

DynamicHMCModels.autocor!

DynamicHMCModels.autocorplot

DynamicHMCModels.autocorplot!

DynamicHMCModels.autocov

DynamicHMCModels.autocov!

DynamicHMCModels.aweights

DynamicHMCModels.axpby!

DynamicHMCModels.axpy!

DynamicHMCModels.backend

DynamicHMCModels.backend_name

DynamicHMCModels.backend_object

DynamicHMCModels.backends

DynamicHMCModels.bar

DynamicHMCModels.bar!

DynamicHMCModels.barh

DynamicHMCModels.barh!

DynamicHMCModels.barhist

DynamicHMCModels.barhist!

DynamicHMCModels.base_color_type

DynamicHMCModels.base_colorant_type

DynamicHMCModels.bbox

DynamicHMCModels.betaccdf

DynamicHMCModels.betacdf

DynamicHMCModels.betainvccdf

DynamicHMCModels.betainvcdf

DynamicHMCModels.betainvlogccdf

DynamicHMCModels.betainvlogcdf

DynamicHMCModels.betalogccdf

DynamicHMCModels.betalogcdf

DynamicHMCModels.betalogpdf

DynamicHMCModels.betapdf

DynamicHMCModels.bic

DynamicHMCModels.binaryentropy

DynamicHMCModels.binomccdf

DynamicHMCModels.binomcdf

DynamicHMCModels.binominvccdf

DynamicHMCModels.binominvcdf

DynamicHMCModels.binominvlogccdf

DynamicHMCModels.binominvlogcdf

DynamicHMCModels.binomlogccdf

DynamicHMCModels.binomlogcdf

DynamicHMCModels.binomlogpdf

DynamicHMCModels.binompdf

DynamicHMCModels.blue

DynamicHMCModels.boxplot

DynamicHMCModels.boxplot!

DynamicHMCModels.brush

DynamicHMCModels.bunchkaufman

DynamicHMCModels.bunchkaufman!

DynamicHMCModels.by

DynamicHMCModels.canonform

DynamicHMCModels.categorical

DynamicHMCModels.categorical!

DynamicHMCModels.ccdf

DynamicHMCModels.ccolor

DynamicHMCModels.cdensity

DynamicHMCModels.cdensity!

DynamicHMCModels.cdf

DynamicHMCModels.center

DynamicHMCModels.cf

DynamicHMCModels.cgf

DynamicHMCModels.cgrad

DynamicHMCModels.cgradients

DynamicHMCModels.chisqccdf

DynamicHMCModels.chisqcdf

DynamicHMCModels.chisqinvccdf

DynamicHMCModels.chisqinvcdf

DynamicHMCModels.chisqinvlogccdf

DynamicHMCModels.chisqinvlogcdf

DynamicHMCModels.chisqlogccdf

DynamicHMCModels.chisqlogcdf

DynamicHMCModels.chisqlogpdf

DynamicHMCModels.chisqpdf

DynamicHMCModels.cholesky

DynamicHMCModels.cholesky!

DynamicHMCModels.cie_color_match

DynamicHMCModels.circvar

DynamicHMCModels.clibraries

DynamicHMCModels.clibrary

DynamicHMCModels.closeall

DynamicHMCModels.coalesce

DynamicHMCModels.coef

DynamicHMCModels.coefnames

DynamicHMCModels.coeftable

DynamicHMCModels.color

DynamicHMCModels.color_type

DynamicHMCModels.coloralpha

DynamicHMCModels.colordiff

DynamicHMCModels.colormap

DynamicHMCModels.colormatch

DynamicHMCModels.colwise

DynamicHMCModels.combine

DynamicHMCModels.comp1

DynamicHMCModels.comp2

DynamicHMCModels.comp3

DynamicHMCModels.compact

DynamicHMCModels.competerank

DynamicHMCModels.complete_cases

DynamicHMCModels.complete_cases!

DynamicHMCModels.completecases

DynamicHMCModels.component

DynamicHMCModels.components

DynamicHMCModels.componentwise_logpdf

DynamicHMCModels.componentwise_pdf

DynamicHMCModels.compress

DynamicHMCModels.concentration

DynamicHMCModels.cond

DynamicHMCModels.condskeel

DynamicHMCModels.confint

DynamicHMCModels.contour

DynamicHMCModels.contour!

DynamicHMCModels.contour3d

DynamicHMCModels.contour3d!

DynamicHMCModels.contourf

DynamicHMCModels.contourf!

DynamicHMCModels.convert

DynamicHMCModels.coords

DynamicHMCModels.copy_transpose!

DynamicHMCModels.copyto!

DynamicHMCModels.cor

DynamicHMCModels.cor2cov

DynamicHMCModels.corkendall

DynamicHMCModels.corner

DynamicHMCModels.corner!

DynamicHMCModels.cornerplot

DynamicHMCModels.cornerplot!

DynamicHMCModels.corrplot

DynamicHMCModels.corrplot!

DynamicHMCModels.corspearman

DynamicHMCModels.counteq

DynamicHMCModels.countmap

DynamicHMCModels.countne

DynamicHMCModels.counts

DynamicHMCModels.cov

DynamicHMCModels.cov2cor

DynamicHMCModels.cquantile

DynamicHMCModels.cross

DynamicHMCModels.crosscor

DynamicHMCModels.crosscor!

DynamicHMCModels.crosscov

DynamicHMCModels.crosscov!

DynamicHMCModels.crossentropy

DynamicHMCModels.cumulant

DynamicHMCModels.current

DynamicHMCModels.curve_points

DynamicHMCModels.curves

DynamicHMCModels.curves!

DynamicHMCModels.cut

DynamicHMCModels.cvec

DynamicHMCModels.dataset

DynamicHMCModels.dataviewer

DynamicHMCModels.decompress

DynamicHMCModels.default

DynamicHMCModels.default_cgrad

DynamicHMCModels.delete!

DynamicHMCModels.deletecols!

DynamicHMCModels.deleterows!

DynamicHMCModels.denserank

DynamicHMCModels.density

DynamicHMCModels.density!

DynamicHMCModels.densityplot

DynamicHMCModels.describe

DynamicHMCModels.deserialize

DynamicHMCModels.det

DynamicHMCModels.deuteranopic

DynamicHMCModels.deviance

DynamicHMCModels.diag

DynamicHMCModels.diagind

DynamicHMCModels.diagm

DynamicHMCModels.dim

DynamicHMCModels.dimension

DynamicHMCModels.disallowmissing

DynamicHMCModels.disallowmissing!

DynamicHMCModels.discretediag

DynamicHMCModels.distinguishable_colors

DynamicHMCModels.diverging_palette

DynamicHMCModels.dof

DynamicHMCModels.dof_residual

DynamicHMCModels.dot

DynamicHMCModels.droplevels!

DynamicHMCModels.dropmissing

DynamicHMCModels.dropmissing!

DynamicHMCModels.ea_histogram

DynamicHMCModels.ea_histogram!

DynamicHMCModels.ecdf

DynamicHMCModels.effective_sample_size

DynamicHMCModels.eigen

DynamicHMCModels.eigen!

DynamicHMCModels.eigmax

DynamicHMCModels.eigmin

DynamicHMCModels.eigvals

DynamicHMCModels.eigvals!

DynamicHMCModels.eigvecs

DynamicHMCModels.eltypes

DynamicHMCModels.entropy

DynamicHMCModels.ess_factor_estimate

DynamicHMCModels.estimate

DynamicHMCModels.eval

DynamicHMCModels.expected_logdet

DynamicHMCModels.factorize

DynamicHMCModels.failprob

DynamicHMCModels.fdistccdf

DynamicHMCModels.fdistcdf

DynamicHMCModels.fdistinvccdf

DynamicHMCModels.fdistinvcdf

DynamicHMCModels.fdistinvlogccdf

DynamicHMCModels.fdistinvlogcdf

DynamicHMCModels.fdistlogccdf

DynamicHMCModels.fdistlogcdf

DynamicHMCModels.fdistlogpdf

DynamicHMCModels.fdistpdf

DynamicHMCModels.findat

DynamicHMCModels.fit

DynamicHMCModels.fit!

DynamicHMCModels.fit_map

DynamicHMCModels.fit_map!

DynamicHMCModels.fit_mle

DynamicHMCModels.fit_mle!

DynamicHMCModels.fitted

DynamicHMCModels.font

DynamicHMCModels.fourinvπ

DynamicHMCModels.fourπ

DynamicHMCModels.frame

DynamicHMCModels.freecumulant

DynamicHMCModels.fweights

DynamicHMCModels.gammaccdf

DynamicHMCModels.gammacdf

DynamicHMCModels.gammainvccdf

DynamicHMCModels.gammainvcdf

DynamicHMCModels.gammainvlogccdf

DynamicHMCModels.gammainvlogcdf

DynamicHMCModels.gammalogccdf

DynamicHMCModels.gammalogcdf

DynamicHMCModels.gammalogpdf

DynamicHMCModels.gammapdf

DynamicHMCModels.gamutmax

DynamicHMCModels.gamutmin

DynamicHMCModels.gelmandiag

DynamicHMCModels.generate

DynamicHMCModels.generate_d — Function.

generate_d

Utility function to generate all notebooks and chapters from scripts in the scripts directory.

Method

generate_d(sd = script_dict_d)

Required arguments

None, all notebooks/.. and chapters/.. files are regenerated.

generate

Generate notebooks and scripts in a single chapter.

Method

generate(chapter::AbstractString)

Required arguments

Generate notebooks and scripts in a single chapter, e.g. generate("04")

generate

Generate a single notebook and script

Method

generate(chapter::AbstractString, file::AbstractString)

Required arguments

Generate notebook and script file in chapter, e.g. generate("04", "m4.1d.jl") or generate("04/m4.1d.jl")

DynamicHMCModels.genmean

DynamicHMCModels.genvar

DynamicHMCModels.geomean

DynamicHMCModels.get_acceptance_rate

DynamicHMCModels.get_color_palette

DynamicHMCModels.get_depth

DynamicHMCModels.get_neg_energy

DynamicHMCModels.get_parent

DynamicHMCModels.get_position

DynamicHMCModels.get_position_matrix

DynamicHMCModels.get_steps

DynamicHMCModels.get_termination

DynamicHMCModels.get_transformation

DynamicHMCModels.getindex

DynamicHMCModels.gewekediag

DynamicHMCModels.gif

DynamicHMCModels.givens

DynamicHMCModels.gkldiv

DynamicHMCModels.gr

DynamicHMCModels.gradlogpdf

DynamicHMCModels.gray

DynamicHMCModels.green

DynamicHMCModels.grid

DynamicHMCModels.groupby

DynamicHMCModels.groupedbar

DynamicHMCModels.groupedbar!

DynamicHMCModels.groupindices

DynamicHMCModels.groupvars

DynamicHMCModels.gui

DynamicHMCModels.halfπ

DynamicHMCModels.harmmean

DynamicHMCModels.hasfinitesupport

DynamicHMCModels.hdf5

DynamicHMCModels.head

DynamicHMCModels.heatmap

DynamicHMCModels.heatmap!

DynamicHMCModels.heideldiag

DynamicHMCModels.hessenberg

DynamicHMCModels.hessenberg!

DynamicHMCModels.hex

DynamicHMCModels.hexbin

DynamicHMCModels.hexbin!

DynamicHMCModels.hist

DynamicHMCModels.histogram

DynamicHMCModels.histogram!

DynamicHMCModels.histogram2d

DynamicHMCModels.histogram2d!

DynamicHMCModels.histogramplot

DynamicHMCModels.hline

DynamicHMCModels.hline!

DynamicHMCModels.hspan

DynamicHMCModels.hspan!

DynamicHMCModels.hyperccdf

DynamicHMCModels.hypercdf

DynamicHMCModels.hyperinvccdf

DynamicHMCModels.hyperinvcdf

DynamicHMCModels.hyperinvlogccdf

DynamicHMCModels.hyperinvlogcdf

DynamicHMCModels.hyperlogccdf

DynamicHMCModels.hyperlogcdf

DynamicHMCModels.hyperlogpdf

DynamicHMCModels.hyperpdf

DynamicHMCModels.include

DynamicHMCModels.indexmap

DynamicHMCModels.indicatormat

DynamicHMCModels.informationmatrix

DynamicHMCModels.inline

DynamicHMCModels.insert!

DynamicHMCModels.insertcols!

DynamicHMCModels.inspectdr

DynamicHMCModels.insupport

DynamicHMCModels.inv2π

DynamicHMCModels.inv4π

DynamicHMCModels.invcov

DynamicHMCModels.inverse

DynamicHMCModels.inverse!

DynamicHMCModels.inverse_eltype

DynamicHMCModels.inverse_rle

DynamicHMCModels.invisible

DynamicHMCModels.invlogccdf

DynamicHMCModels.invlogcdf

DynamicHMCModels.invscale

DynamicHMCModels.invsoftplus

DynamicHMCModels.invsqrt2

DynamicHMCModels.invsqrt2π

DynamicHMCModels.invπ

DynamicHMCModels.iqr

DynamicHMCModels.isbounded

DynamicHMCModels.isdark

DynamicHMCModels.isdiag

DynamicHMCModels.isfitted

DynamicHMCModels.ishermitian

DynamicHMCModels.isleptokurtic

DynamicHMCModels.islinear

DynamicHMCModels.islowerbounded

DynamicHMCModels.ismesokurtic

DynamicHMCModels.ismissing

DynamicHMCModels.isordered

DynamicHMCModels.isplatykurtic

DynamicHMCModels.isposdef

DynamicHMCModels.isposdef!

DynamicHMCModels.isprobvec

DynamicHMCModels.issuccess

DynamicHMCModels.issymmetric

DynamicHMCModels.istril

DynamicHMCModels.istriu

DynamicHMCModels.isupperbounded

DynamicHMCModels.iter_segments

DynamicHMCModels.kde

DynamicHMCModels.keys

DynamicHMCModels.kldivergence

DynamicHMCModels.kron

DynamicHMCModels.kurtosis

DynamicHMCModels.ldiv!

DynamicHMCModels.ldlt

DynamicHMCModels.ldlt!

DynamicHMCModels.length

DynamicHMCModels.levels

DynamicHMCModels.levels!

DynamicHMCModels.levelsmap

DynamicHMCModels.leverage

DynamicHMCModels.link

DynamicHMCModels.lmul!

DynamicHMCModels.location

DynamicHMCModels.location!

DynamicHMCModels.log1mexp

DynamicHMCModels.log1pexp

DynamicHMCModels.log1pmx

DynamicHMCModels.log1psq

DynamicHMCModels.log2mexp

DynamicHMCModels.log2π

DynamicHMCModels.log4π

DynamicHMCModels.logabsdet

DynamicHMCModels.logaddexp

DynamicHMCModels.logccdf

DynamicHMCModels.logcdf

DynamicHMCModels.logdensity

DynamicHMCModels.logdet

DynamicHMCModels.logdetcov

DynamicHMCModels.logexpm1

DynamicHMCModels.loghalf

DynamicHMCModels.logistic

DynamicHMCModels.logit

DynamicHMCModels.logjac_forwarddiff

DynamicHMCModels.loglikelihood

DynamicHMCModels.logmvgamma

DynamicHMCModels.logmxp1

DynamicHMCModels.logpdf

DynamicHMCModels.logpdf!

DynamicHMCModels.logsumexp

DynamicHMCModels.logtwo

DynamicHMCModels.logπ

DynamicHMCModels.lowrankdowndate

DynamicHMCModels.lowrankdowndate!

DynamicHMCModels.lowrankupdate

DynamicHMCModels.lowrankupdate!

DynamicHMCModels.lq

DynamicHMCModels.lq!

DynamicHMCModels.lstirling_asym

DynamicHMCModels.lu

DynamicHMCModels.lu!

DynamicHMCModels.lyap

DynamicHMCModels.mad

DynamicHMCModels.mad!

DynamicHMCModels.map

DynamicHMCModels.mapc

DynamicHMCModels.mapcols

DynamicHMCModels.mapreducec

DynamicHMCModels.marginalhist

DynamicHMCModels.marginalhist!

DynamicHMCModels.maxad

DynamicHMCModels.mcmc

DynamicHMCModels.mcmc_adapting_ϵ

DynamicHMCModels.mean

DynamicHMCModels.mean_and_cov

DynamicHMCModels.mean_and_std

DynamicHMCModels.mean_and_var

DynamicHMCModels.meanad

DynamicHMCModels.meandir

DynamicHMCModels.meanform

DynamicHMCModels.meanlogx

DynamicHMCModels.meanplot

DynamicHMCModels.meanplot!

DynamicHMCModels.meanresponse

DynamicHMCModels.median

DynamicHMCModels.melt

DynamicHMCModels.meltdf

DynamicHMCModels.merge!

DynamicHMCModels.mgf

DynamicHMCModels.middle

DynamicHMCModels.midpoints

DynamicHMCModels.missing

DynamicHMCModels.missings

DynamicHMCModels.mixeddensity

DynamicHMCModels.mixeddensity!

DynamicHMCModels.mixeddensityplot

DynamicHMCModels.mode

DynamicHMCModels.model_response

DynamicHMCModels.modelmatrix

DynamicHMCModels.modes

DynamicHMCModels.moment

DynamicHMCModels.mov

DynamicHMCModels.mp4

DynamicHMCModels.msd

DynamicHMCModels.mss

DynamicHMCModels.mul!

DynamicHMCModels.names!

DynamicHMCModels.nbetaccdf

DynamicHMCModels.nbetacdf

DynamicHMCModels.nbetainvccdf

DynamicHMCModels.nbetainvcdf

DynamicHMCModels.nbetainvlogccdf

DynamicHMCModels.nbetainvlogcdf

DynamicHMCModels.nbetalogccdf

DynamicHMCModels.nbetalogcdf

DynamicHMCModels.nbetalogpdf

DynamicHMCModels.nbetapdf

DynamicHMCModels.nbinomccdf

DynamicHMCModels.nbinomcdf

DynamicHMCModels.nbinominvccdf

DynamicHMCModels.nbinominvcdf

DynamicHMCModels.nbinominvlogccdf

DynamicHMCModels.nbinominvlogcdf

DynamicHMCModels.nbinomlogccdf

DynamicHMCModels.nbinomlogcdf

DynamicHMCModels.nbinomlogpdf

DynamicHMCModels.nbinompdf

DynamicHMCModels.ncategories

DynamicHMCModels.nchisqccdf

DynamicHMCModels.nchisqcdf

DynamicHMCModels.nchisqinvccdf

DynamicHMCModels.nchisqinvcdf

DynamicHMCModels.nchisqinvlogccdf

DynamicHMCModels.nchisqinvlogcdf

DynamicHMCModels.nchisqlogccdf

DynamicHMCModels.nchisqlogcdf

DynamicHMCModels.nchisqlogpdf

DynamicHMCModels.nchisqpdf

DynamicHMCModels.ncol

DynamicHMCModels.ncomponents

DynamicHMCModels.nfdistccdf

DynamicHMCModels.nfdistcdf

DynamicHMCModels.nfdistinvccdf

DynamicHMCModels.nfdistinvcdf

DynamicHMCModels.nfdistinvlogccdf

DynamicHMCModels.nfdistinvlogcdf

DynamicHMCModels.nfdistlogccdf

DynamicHMCModels.nfdistlogcdf

DynamicHMCModels.nfdistlogpdf

DynamicHMCModels.nfdistpdf

DynamicHMCModels.nobs

DynamicHMCModels.nonunique

DynamicHMCModels.norepeats

DynamicHMCModels.norm

DynamicHMCModels.normalize

DynamicHMCModels.normalize!

DynamicHMCModels.normccdf

DynamicHMCModels.normcdf

DynamicHMCModels.norminvccdf

DynamicHMCModels.norminvcdf

DynamicHMCModels.norminvlogccdf

DynamicHMCModels.norminvlogcdf

DynamicHMCModels.normlogccdf

DynamicHMCModels.normlogcdf

DynamicHMCModels.normlogpdf

DynamicHMCModels.normpdf

DynamicHMCModels.nquantile

DynamicHMCModels.nrow

DynamicHMCModels.nsamples

DynamicHMCModels.ntdistccdf

DynamicHMCModels.ntdistcdf

DynamicHMCModels.ntdistinvccdf

DynamicHMCModels.ntdistinvcdf

DynamicHMCModels.ntdistinvlogccdf

DynamicHMCModels.ntdistinvlogcdf

DynamicHMCModels.ntdistlogccdf

DynamicHMCModels.ntdistlogcdf

DynamicHMCModels.ntdistlogpdf

DynamicHMCModels.ntdistpdf

DynamicHMCModels.ntrials

DynamicHMCModels.nullable!

DynamicHMCModels.nulldeviance

DynamicHMCModels.nullloglikelihood

DynamicHMCModels.nullspace

DynamicHMCModels.ohlc

DynamicHMCModels.ohlc!

DynamicHMCModels.opnorm

DynamicHMCModels.optimize_datetime_ticks

DynamicHMCModels.optimize_ticks

DynamicHMCModels.order

DynamicHMCModels.ordered

DynamicHMCModels.ordered!

DynamicHMCModels.ordinalrank

DynamicHMCModels.ordschur

DynamicHMCModels.ordschur!

DynamicHMCModels.pacf

DynamicHMCModels.pacf!

DynamicHMCModels.palette

DynamicHMCModels.params

DynamicHMCModels.params!

DynamicHMCModels.partialcor

DynamicHMCModels.partype

DynamicHMCModels.passmissing

DynamicHMCModels.path3d

DynamicHMCModels.path3d!

DynamicHMCModels.pdf

DynamicHMCModels.percentile

DynamicHMCModels.permutecols!

DynamicHMCModels.pgfplots

DynamicHMCModels.pie

DynamicHMCModels.pie!

DynamicHMCModels.pinv

DynamicHMCModels.plot

DynamicHMCModels.plot!

DynamicHMCModels.plot3d

DynamicHMCModels.plot3d!

DynamicHMCModels.plot_color

DynamicHMCModels.plotarea

DynamicHMCModels.plotattr

DynamicHMCModels.plotly

DynamicHMCModels.plotlyjs

DynamicHMCModels.plots_heatmap

DynamicHMCModels.plots_heatmap!

DynamicHMCModels.png

DynamicHMCModels.poisccdf

DynamicHMCModels.poiscdf

DynamicHMCModels.poisinvccdf

DynamicHMCModels.poisinvcdf

DynamicHMCModels.poisinvlogccdf

DynamicHMCModels.poisinvlogcdf

DynamicHMCModels.poislogccdf

DynamicHMCModels.poislogcdf

DynamicHMCModels.poislogpdf

DynamicHMCModels.poispdf

DynamicHMCModels.pool

DynamicHMCModels.pool!

DynamicHMCModels.portfoliocomposition

DynamicHMCModels.portfoliocomposition!

DynamicHMCModels.potential_scale_reduction

DynamicHMCModels.predict

DynamicHMCModels.predict!

DynamicHMCModels.probs

DynamicHMCModels.probval

DynamicHMCModels.proportionmap

DynamicHMCModels.proportions

DynamicHMCModels.protanopic

DynamicHMCModels.psnr

DynamicHMCModels.pweights

DynamicHMCModels.pyplot

DynamicHMCModels.qqbuild

DynamicHMCModels.qqnorm

DynamicHMCModels.qqnorm!

DynamicHMCModels.qqplot

DynamicHMCModels.qqplot!

DynamicHMCModels.qr

DynamicHMCModels.qr!

DynamicHMCModels.quantile

DynamicHMCModels.quartπ

DynamicHMCModels.quiver

DynamicHMCModels.quiver!

DynamicHMCModels.r2

DynamicHMCModels.rafterydiag

DynamicHMCModels.random_arg

DynamicHMCModels.random_value

DynamicHMCModels.rank

DynamicHMCModels.rate

DynamicHMCModels.rdiv!

DynamicHMCModels.readdlm

DynamicHMCModels.readtable

DynamicHMCModels.recode

DynamicHMCModels.recode!

DynamicHMCModels.reconstruct

DynamicHMCModels.red

DynamicHMCModels.reducec

DynamicHMCModels.reject_logdensity

DynamicHMCModels.rel_path

DynamicHMCModels.rel_path_d — Function.

relpathd

Relative path using the DynamicHMCModels src/ directory.

Example to get access to the data subdirectory

rel_path_d("..", "data")

DynamicHMCModels.rename

DynamicHMCModels.rename!

DynamicHMCModels.renyientropy

DynamicHMCModels.residuals

DynamicHMCModels.response

DynamicHMCModels.rgb_string

DynamicHMCModels.rgba_string

DynamicHMCModels.rle

DynamicHMCModels.rmsd

DynamicHMCModels.rmul!

DynamicHMCModels.rotate

DynamicHMCModels.rotate!

DynamicHMCModels.rss

DynamicHMCModels.r²

DynamicHMCModels.sample

DynamicHMCModels.sample!

DynamicHMCModels.sample_cov

DynamicHMCModels.samplepair

DynamicHMCModels.sampler

DynamicHMCModels.savefig

DynamicHMCModels.scale

DynamicHMCModels.scale!

DynamicHMCModels.scatter

DynamicHMCModels.scatter!

DynamicHMCModels.scatter3d

DynamicHMCModels.scatter3d!

DynamicHMCModels.scatterhist

DynamicHMCModels.scatterhist!

DynamicHMCModels.scattermat

DynamicHMCModels.schur

DynamicHMCModels.schur!

DynamicHMCModels.score

DynamicHMCModels.script_dict

DynamicHMCModels.script_dict_d

DynamicHMCModels.scriptentry

DynamicHMCModels.sem

DynamicHMCModels.sequential_palette

DynamicHMCModels.serialize

DynamicHMCModels.set_theme

DynamicHMCModels.setindex!

DynamicHMCModels.shape

DynamicHMCModels.shape_coords

DynamicHMCModels.show

DynamicHMCModels.showall

DynamicHMCModels.showcols

DynamicHMCModels.showgradient

DynamicHMCModels.showgradient!

DynamicHMCModels.showlibrary

DynamicHMCModels.showlibrary!

DynamicHMCModels.showtheme

DynamicHMCModels.showtheme!

DynamicHMCModels.skewness

DynamicHMCModels.softmax

DynamicHMCModels.softmax!

DynamicHMCModels.softplus

DynamicHMCModels.span

DynamicHMCModels.spy

DynamicHMCModels.spy!

DynamicHMCModels.sqL2dist

DynamicHMCModels.sqmahal

DynamicHMCModels.sqmahal!

DynamicHMCModels.sqrt2

DynamicHMCModels.sqrt2π

DynamicHMCModels.sqrt3

DynamicHMCModels.sqrt4π

DynamicHMCModels.sqrthalfπ

DynamicHMCModels.sqrtπ

DynamicHMCModels.src_path_d

DynamicHMCModels.srdistccdf

DynamicHMCModels.srdistcdf

DynamicHMCModels.srdistinvccdf

DynamicHMCModels.srdistinvcdf

DynamicHMCModels.srdistinvlogccdf

DynamicHMCModels.srdistinvlogcdf

DynamicHMCModels.srdistlogccdf

DynamicHMCModels.srdistlogcdf

DynamicHMCModels.stack

DynamicHMCModels.stackdf

DynamicHMCModels.std

DynamicHMCModels.stderror

DynamicHMCModels.stdlogx

DynamicHMCModels.stephist

DynamicHMCModels.stephist!

DynamicHMCModels.sticks

DynamicHMCModels.sticks!

DynamicHMCModels.stroke

DynamicHMCModels.sub

DynamicHMCModels.succprob

DynamicHMCModels.suffstats

DynamicHMCModels.summarystats

DynamicHMCModels.support

DynamicHMCModels.surface

DynamicHMCModels.surface!

DynamicHMCModels.svd

DynamicHMCModels.svd!

DynamicHMCModels.svdvals

DynamicHMCModels.svdvals!

DynamicHMCModels.sylvester

DynamicHMCModels.tail

DynamicHMCModels.tdistccdf

DynamicHMCModels.tdistcdf

DynamicHMCModels.tdistinvccdf

DynamicHMCModels.tdistinvcdf

DynamicHMCModels.tdistinvlogccdf

DynamicHMCModels.tdistinvlogcdf

DynamicHMCModels.tdistlogccdf

DynamicHMCModels.tdistlogcdf

DynamicHMCModels.tdistlogpdf

DynamicHMCModels.tdistpdf

DynamicHMCModels.test_distr

DynamicHMCModels.test_examples

DynamicHMCModels.test_samples

DynamicHMCModels.testhist

DynamicHMCModels.testhist!

DynamicHMCModels.text

DynamicHMCModels.theme

DynamicHMCModels.tiedrank

DynamicHMCModels.title!

DynamicHMCModels.to_array

DynamicHMCModels.to_tuple

DynamicHMCModels.totalvar

DynamicHMCModels.tr

DynamicHMCModels.traceplot

DynamicHMCModels.traceplot!

DynamicHMCModels.transform

DynamicHMCModels.transform_and_logjac

DynamicHMCModels.transform_logdensity

DynamicHMCModels.translate

DynamicHMCModels.translate!

DynamicHMCModels.transpose

DynamicHMCModels.transpose!

DynamicHMCModels.tril

DynamicHMCModels.tril!

DynamicHMCModels.trim

DynamicHMCModels.trim!

DynamicHMCModels.trimvar

DynamicHMCModels.tritanopic

DynamicHMCModels.triu

DynamicHMCModels.triu!

DynamicHMCModels.twinx

DynamicHMCModels.twoinvπ

DynamicHMCModels.twoπ

DynamicHMCModels.type2dict

DynamicHMCModels.uncompact

DynamicHMCModels.unicodeplots

DynamicHMCModels.unique!

DynamicHMCModels.unstack

DynamicHMCModels.value_and_logjac_forwarddiff

DynamicHMCModels.values

DynamicHMCModels.var

DynamicHMCModels.variation

DynamicHMCModels.varlogx

DynamicHMCModels.vcat

DynamicHMCModels.vcov

DynamicHMCModels.violin

DynamicHMCModels.violin!

DynamicHMCModels.vline

DynamicHMCModels.vline!

DynamicHMCModels.vspan

DynamicHMCModels.vspan!

DynamicHMCModels.weighted_color_mean

DynamicHMCModels.weights

DynamicHMCModels.whitebalance

DynamicHMCModels.winsor

DynamicHMCModels.winsor!

DynamicHMCModels.wireframe

DynamicHMCModels.wireframe!

DynamicHMCModels.with

DynamicHMCModels.wmean

DynamicHMCModels.wmean!

DynamicHMCModels.wmedian

DynamicHMCModels.wquantile

DynamicHMCModels.wrap

DynamicHMCModels.writedlm

DynamicHMCModels.writetable

DynamicHMCModels.wsample

DynamicHMCModels.wsample!

DynamicHMCModels.wsum

DynamicHMCModels.wsum!

DynamicHMCModels.xaxis!

DynamicHMCModels.xflip!

DynamicHMCModels.xgrid!

DynamicHMCModels.xlabel!

DynamicHMCModels.xlims

DynamicHMCModels.xlims!

DynamicHMCModels.xlogx

DynamicHMCModels.xlogy

DynamicHMCModels.xticks!

DynamicHMCModels.xyY

DynamicHMCModels.xyYA

DynamicHMCModels.yaxis!

DynamicHMCModels.yflip!

DynamicHMCModels.ygrid!

DynamicHMCModels.ylabel!

DynamicHMCModels.ylims

DynamicHMCModels.ylims!

DynamicHMCModels.yticks!

DynamicHMCModels.zlims

DynamicHMCModels.zlims!

DynamicHMCModels.zscore

DynamicHMCModels.zscore!

DynamicHMCModels.×

DynamicHMCModels.∞

DynamicHMCModels.⋅