Docstrings

LaTeXTabulars.CMidRule — Type.

CMidRule([wd], [trim], left, right)

Will be printed as \cmidrule[wd](trim)[left-right]. When wd or trim is nothing, it is omitted. Use with the booktabs LaTeX package.

LaTeXTabulars.LaTeXTabulars

LaTeXTabulars.LongTable

LaTeXTabulars.MultiColumn — Type.

MultiColumn(n, pos, cell)

For \multicolumn{n}{pos}{cell}. Use the symbols :l, :c, :r for pos.

LaTeXTabulars.Rule — Type.

Rule()
Rule(kind)

Horizontal rule. The kind of the rule is specified by a symbol, which will generally be printed as \KINDrule for rules in booktabs, eg Rule(:top) prints \toprule. To obtain a \hline, use Rule{:h}.

LaTeXTabulars.Tabular — Type.

Tabular(cols)

For the LaTeX environment \begin{tabular}{cols} ... \end{tabular}.

LaTeXTabulars.TabularLike

LaTeXTabulars.eval

LaTeXTabulars.include

LaTeXTabulars.latex_cell — Function.

latex_cell(io, cell)

Print a the contents of cell to io as LaTeX.

!!! NOTE

Methods are defined for some specific types, but if you want full control
(eg rounding), use an `<: AbstractString`, eg `String` or `LaTeXString`.

LaTeXTabulars.latex_env_begin

LaTeXTabulars.latex_env_end

LaTeXTabulars.latex_line

LaTeXTabulars.latex_tabular — Function.

latex_tabular(io, t, lines)

Print lines to io as a LaTeX using the given environment.

Each line in lines can be

a rule-like object, eg [Rule] or [CMidRule],

an iterable (eg AbstractVector) of cells,

a Tuple, which is treated as multiple lines (“splat” in place), which is useful for functions that generate lines with associated rules, or multiple CMidRules,

a matrix, each row of which is treated as a line.

See latex_cell for the kinds of cell supported (particularly MultiColumn, but for full formatting control, use an String or LaTeXString for cells.

latex_tabular(?, t, lines)

LaTeX output as a string.

latex_tabular(filename, t, lines)

Write a tabular-like LaTeX environment to filename, which is overwritten if it already exists.