Docstrings

DataFitting.@code_ndim

DataFitting.AbstractCartesianDomain

DataFitting.AbstractComponent

DataFitting.AbstractComponentData

DataFitting.AbstractCounts

DataFitting.AbstractData

DataFitting.AbstractDomain

DataFitting.AbstractLinearDomain

DataFitting.AbstractMeasures

DataFitting.AbstractMinimizer

DataFitting.CartesianDomain

DataFitting.CartesianDomain_2D

DataFitting.CompiledExpression

DataFitting.ComponentEvaluation

DataFitting.Counts

DataFitting.Counts_1D

DataFitting.Counts_2D

DataFitting.DataFitting

DataFitting.Domain

DataFitting.Domain_1D

DataFitting.Domain_2D

DataFitting.FitParameter

DataFitting.FitResult

DataFitting.FuncWrap

DataFitting.FuncWrap_cdata

DataFitting.HashVector

DataFitting.Measures

DataFitting.Measures_1D

DataFitting.Measures_2D

DataFitting.Meta_parse

DataFitting.Minimizer

DataFitting.Model — Type.

Model

Constructor for the Model structure.

Returns a component evaluation.

DataFitting.Parameter

DataFitting.SimpleParam

DataFitting.SimpleParam_cdata

DataFitting.checkID

DataFitting.compdata

DataFitting.domain — Function.

domain

Return the domain associated to a model.

DataFitting.eval

DataFitting.evalcounter — Function.

evalcounter

Return the number of times the model has been evaluated.

evalcounter

Return the number of times a component has been evaluated.

DataFitting.evaluate! — Function.

evaluate!

Evaluate the model(s)

DataFitting.fit! — Function.

fit!

Fit a model against data, using the specified minimizer.

DataFitting.flatten

DataFitting.getaxisextrema

DataFitting.getaxismax

DataFitting.getaxismin

DataFitting.getparams

DataFitting.getsetparams!

DataFitting.include

DataFitting.keyToIndex

DataFitting.minimize

DataFitting.nparams — Function.

nparams

Returns number of params in a component.

DataFitting.oldver

DataFitting.prepare! — Function.

prepare!

Recompile the expressions after one (or more) parameter expressions have been changed

prepare!

Prepare a model to be evaluated on the given domain, with the given mathematical expression to join the components.

DataFitting.resetcounters! — Function.

resetcounters!

Reset model and components evaluation counters.

DataFitting.setcompvalue! — Function.

setcompvalue!

Set a component value in the model

DataFitting.support_param_limits

DataFitting.test_component

DataFitting.values