Docstrings

Flatten.@flattenable

Flatten.@reflattenable

Flatten.Flatten

Flatten.__init__

Flatten.default_combiner

Flatten.eval

Flatten.fieldname_meta

Flatten.fieldnameflatten

Flatten.fieldparent_meta

Flatten.fieldparenttype_meta

Flatten.fieldtype_meta

Flatten.fieldtypeflatten

Flatten.flatten — Function.

Flattening. Flattens a nested type to a Tuple or Vector

Flatten.flatten_expr

Flatten.flatten_inner

Flatten.flattenable

Flatten.include

Flatten.metaflatten — Function.

Tag flattening. Flattens data attached to a field by methods of a passed in function

Flatten.metaflatten_expr

Flatten.metaflatten_inner

Flatten.nested

Flatten.parentflatten

Flatten.parenttypeflatten

Flatten.reconstruct — Function.

Reconstruct an object from partial Tuple or Vector data and another object

Flatten.reconstruct_combiner

Flatten.reconstruct_expr

Flatten.reconstruct_inner

Flatten.retype — Function.

Retype an object from partial Tuple or Vector data and another object

Flatten.retype_combiner

Flatten.retype_expr

Flatten.retype_inner

Flatten.update! — Function.

Update a mutable object with partial Tuple or Vector data

Flatten.update_combiner

Flatten.update_expr

Flatten.update_inner