Docstrings

Graph500.Graph500

Graph500.Graph500Results

Graph500.driver

Graph500.eval

Graph500.get_min_type — Function.

get_min_type(n)

Return the minimum Integer type which can contain n

Graph500.get_next_type — Function.

get_next_type(T)

Return a data with a Integer Datatype twice the size of T

Graph500.include

Graph500.kernel_1 — Function.

kernel_1(ij)

Create Graph from edge list ij based on kernel1 benchmark(http://graph500.org/?pageid=12#sec-4) Return LightGraphs.SimpleGraph structure

Graph500.kernel_2

Graph500.key_sampling

Graph500.kronecker_generator

Graph500.output

Graph500.validate