Docstrings

SingularSpectrumAnalysis.SingularSpectrumAnalysis — Module.

SingularSpectrumAnalysis

Simple Usage:

# generate some data
L = 20 # Window length
K = 100
N = K*L; # number of datapoints
t = 1:N; # Time vector
T = 20; # period of main oscillation
y = sin.(2pi/T*t); # Signal
y .+= (0.5sin.(2pi/T*4*t)).^2 # Add another frequency
e = 0.1randn(N); # Add some noise
ys = y+e;
# plot(ys)

USV = hsvd(ys,L) # Perform svd on the trajectory matrix
sigmaplot(USV) # Plot normalized singular values
# logsigmaplot(USV) # Plot singular values
# cumsigmaplot(USV) # Plot cumulative normalized singular values
seasonal_groupings = [1:2, 4:5] # Determine pairs of singular values corresponding to seasonal components
trends = 3 # If some singular value lacks a buddy, this is a trend component
pairplot(USV,seasonal_groupings) # plot phase plots for all seasonal components
yrt, yrs = reconstruct(USV, trends, seasonal_groupings) # Reconstruct the underlying signal without noise, based on all identified components with significant singular values
yr = sum([yrt yrs],2) # Form full reconstruction
plot([y ys yr], lab=["y" "ys" "yr"])

See http://www.jds-online.com/files/JDS-396.pdf for an easy-to-read introduction to SSA

SingularSpectrumAnalysis.cumsigmaplot

SingularSpectrumAnalysis.elementary — Function.

Ui = elementary(USV,I)

Computes the sum UᵢSᵢVᵢ' for all i in I If I is 1:L, the returned matrix is identical to USV'

SingularSpectrumAnalysis.eval

SingularSpectrumAnalysis.hankel — Function.

X = hankel(x,window_size)

Form the trajectory matrix X of size KxL, K = N-L+1 x can be a vector or a matrix for multivariate SSA

SingularSpectrumAnalysis.hankelize — Function.

means = hankelize(X)

Computes a timeseries from an approximate Hankel matrix by diagoal averaging (Hankelization). Note that the returned value is a vector and not a hankel matrix. A hankel matrix is easily obtainable by hankel(hankelize(X), size(X,2))

SingularSpectrumAnalysis.hsvd — Function.

USV = hsvd(y,L)

Form a trajectory hankel matrix from data y and compute svd on this Hankelmatrix

SingularSpectrumAnalysis.include

SingularSpectrumAnalysis.logsigmaplot

SingularSpectrumAnalysis.pairplot — Function.

pairplot(USV, groupings) Usage:

USV = hsvd(data,L)
seasonal_groupings = [1:2, 4:5]
pairplot(USV,seasonal_groupings)

SingularSpectrumAnalysis.reconstruct — Function.

yrt,yrs = reconstruct(USV::SVD, trends, seasonal::AbstractArray)
yr      = reconstruct(USV::SVD, groupings::AbstractArray)

Compute a reconstruction of the time-series based on an SVD object obtained from hsvd and user selected groupings. See also ?SingularSpectrumAnalysis

SingularSpectrumAnalysis.sigmaplot