Docstrings

CoupledFields.AdditiveKernel

CoupledFields.BaseKernel

CoupledFields.CVfn — Function.

CVfn(parm::T, X::T, Y::T, modelfn::Function, kerneltype::DataType; verbose::Bool=true, dcv::Int64=2) where T<:Matrix{Float64}

Cross-validation function

CoupledFields.CompositeKernel

CoupledFields.CoupledFields

CoupledFields.ExponentialKernel

CoupledFields.GaussianKP — Type.

GaussianKP: For the gaussian kernel

CoupledFields.GaussianKernel

CoupledFields.InputSpace — Type.

InputSpace: A type to hold the `X` and `Y` fields of the Input space

InputSpace(X, Y, d, lat): The fields are whitened if d=[d1, d2] is supplied. Area weighting is applied if lat is supplied.

CoupledFields.Kernel

CoupledFields.KernelParameters — Type.

KernelParameters: An abstract type.

All KernelParameters types contain certain parameters which are later passed to internal functions Kf and ∇Kf.

A KernelParameters type is set using e.g. PolynomialKP(X::Matrix{Float64}) or GaussianKP(X::Matrix{Float64}).

CoupledFields.Kf

CoupledFields.LaplacianKernel

CoupledFields.LinearKernel

CoupledFields.ModelObj — Type.

ModelObj: A type to hold statistical model results

Such as the matrices W, R, A, T, where R=XW and T=YA.

CoupledFields.PolynomialKP — Type.

PolynomialKP: For the polynomial kernel

CoupledFields.PolynomialKernel

CoupledFields.RadialBasisKernel

CoupledFields.Rsq_adj — Function.

Rsq_adj(Tx::T, Ty::T, df::Int) where T<:Array{Float64}:

Cross-validation metric

CoupledFields.ScalarProductKernel

CoupledFields.SeparableKernel

CoupledFields.SquaredDistanceKernel

CoupledFields.StandardKernel

CoupledFields.bf — Function.

bf(x::Vector{Float64}, df::Int):

Compute a piecewise linear basis matrix for the vector x.

CoupledFields.cca — Function.

cca(v::Array{Float64}, X::T,Y::T) where T<:Matrix{Float64}:

Regularized Canonical Correlation Analysis using SVD.

CoupledFields.convert

CoupledFields.eval

CoupledFields.gKCCA — Function.

gKCCA(par::Array{Float64}, X::Matrix{Float64}, Y::Matrix{Float64}, kpars::KernelParameters):

Compute the projection matrices and components for gKCCA.

CoupledFields.gradvecfield — Function.

gradvecfield(par::Array{Float64}, X::T, Y::T, kpars::KernelParameters) where T<:Matrix{Float64}:

Compute the gradient vector or gradient matrix at each instance of the X and Y fields, by making use of a kernel feature space.

CoupledFields.include

CoupledFields.init_pairwise

CoupledFields.ismercer

CoupledFields.isnegdef

CoupledFields.isnonnegative

CoupledFields.kernelmatrix

CoupledFields.kernelmatrix!

CoupledFields.kernelrange

CoupledFields.pairwise!

CoupledFields.pairwise_XY!

CoupledFields.pairwise_XtYt!

CoupledFields.phi

CoupledFields.whiten — Function.

whiten(x::Matrix{Float64}, d::Float64; lat=nothing): Whiten matrix

d (0-1) Percentage variance of components to retain.

lat Latitudinal area-weighting.

CoupledFields.∇Kf