Docstrings

ErlangTerm.ATOM

ErlangTerm.ATOM_UTF8

ErlangTerm.BINARY

ErlangTerm.ErlangTerm

ErlangTerm.INTEGER

ErlangTerm.LARGE_TUPLE

ErlangTerm.LIST

ErlangTerm.MAP

ErlangTerm.NEW_FLOAT

ErlangTerm.NIL

ErlangTerm.SMALL_ATOM_UTF8

ErlangTerm.SMALL_INTEGER

ErlangTerm.SMALL_TUPLE

ErlangTerm.STRING

ErlangTerm.VERSION

ErlangTerm._serialize

ErlangTerm.deserialize — Function.

deserialize(io::IO)
deserialize(binary::Array{UInt8})

Deserialize a Julia value encoded in the Erlang external term format from a stream or an array of bytes.

ErlangTerm.eval

ErlangTerm.include

ErlangTerm.serialize — Function.

serialize(io::IO, value)
serialize(value)

Serialize a Julia value to the Erlang external term format.