Docstrings

JSONSchema.@doassert

JSONSchema.@report

JSONSchema.AbstractIssue

JSONSchema.JSONSchema

JSONSchema.OneOfIssue

JSONSchema.Schema — Type.

Schema(sch::String)

Create a schema for document validation. sch should be a String containing a valid JSON.

Schema(sch::Dict)

Create a schema but with sch being a parsed JSON created with JSON.parse() or JSON.parsefile().

Example

julia> myschema = Schema("
  {
    "properties": {
      "foo": {},
      "bar": {}
    },
    "required": ["foo"]
  }
  ")

julia> sch = JSON.parsefile(filepath)
julia> myschema = Schema(sch)

JSONSchema.SingleIssue

JSONSchema.array_asserts

JSONSchema.diagnose — Function.

diagnose(x, s::Schema)

Check that a document x is valid against the Schema s. If valid return nothing, and if not, return a diagnostic String containing a selection of one or more likely causes of failure.

Examples

julia> diagnose( JSON.parse("{ "foo": true }"), myschema)
nothing
julia> diagnose( JSON.parse("{ "bar": 12.5 }"), myschema)
"in [] : required property 'foo' missing"

JSONSchema.eval

JSONSchema.findelement

JSONSchema.findref

JSONSchema.flatten

JSONSchema.getremoteschema

JSONSchema.include

Base.isvalid — Function.

isvalid(x, s::Schema)

Check that a document x is valid against the Schema s.

Examples

julia> myschema = Schema("
  {
    "properties": {
      "foo": {},
      "bar": {}
    },
    "required": ["foo"]
  }
  ")

julia> isvalid( JSON.parse("{ "foo": true }"), myschema)
true
julia> isvalid( JSON.parse("{ "bar": 12.5 }"), myschema)
false

JSONSchema.mkidmap!

JSONSchema.number_asserts

JSONSchema.object_asserts

JSONSchema.resolverefs!

JSONSchema.rmfragment

JSONSchema.singleissuerecap

JSONSchema.string_asserts

JSONSchema.todict

JSONSchema.type_asserts

JSONSchema.unescapeJPath

JSONSchema.updateid

JSONSchema.validate