Docstrings

ExtremeStats.AbstractMaxima — Type.

AbstractMaxima

A collection of maxima (e.g. block maxima)

ExtremeStats.BlockMaxima — Type.

BlockMaxima(x, n)

Maxima obtained by splitting the data x into blocks of size n.

ExtremeStats.ExcessPlot

ExtremeStats.ExtremeStats

ExtremeStats.ParetoPlot

ExtremeStats.PeakOverThreshold — Type.

PeakOverThreshold(x, u)

Maxima obtained by thresholding the data x with threshold u.

ExtremeStats.ReturnPlot

ExtremeStats.eval

ExtremeStats.excessplot

ExtremeStats.excessplot!

ExtremeStats.include

ExtremeStats.meanexcess — Function.

meanexcess(xs, k)

Return mean excess of the data xs using previous k values.

ExtremeStats.paretoplot

ExtremeStats.paretoplot!

ExtremeStats.returnlevels — Function.

returnlevels(xs)

Return periods and levels of data xs.

returnlevels(gev, mmin, mmax; nlevels=50)

Return nlevels periods and levels of generalized extreme value distribution gev with maxima in the interval [mmin,mmax].

ExtremeStats.returnplot

ExtremeStats.returnplot!