Docstrings

StanModels.+

StanModels.-

StanModels.@animate

StanModels.@colorant_str

StanModels.@csv2_str

StanModels.@csv_str

StanModels.@df

StanModels.@gif

StanModels.@layout

StanModels.@load

StanModels.@recipe

StanModels.@save

StanModels.@series

StanModels.@shorthands

StanModels.@tsv_str

StanModels.@userplot

StanModels.@wsv_str

StanModels.ABGR

StanModels.ADIN99

StanModels.ADIN99d

StanModels.ADIN99o

StanModels.AGray

StanModels.AGray32

StanModels.AHSI

StanModels.AHSL

StanModels.AHSV

StanModels.ALCHab

StanModels.ALCHuv

StanModels.ALMS

StanModels.ALab

StanModels.ALuv

StanModels.ARGB

StanModels.ARGB32

StanModels.AXYZ

StanModels.AYCbCr

StanModels.AYIQ

StanModels.AbstractBackend

StanModels.AbstractCategoricalArray

StanModels.AbstractCategoricalMatrix

StanModels.AbstractCategoricalVector

StanModels.AbstractDataFrame

StanModels.AbstractGray

StanModels.AbstractHistogram

StanModels.AbstractLayout

StanModels.AbstractMissingCategoricalArray

StanModels.AbstractMissingCategoricalMatrix

StanModels.AbstractMissingCategoricalVector

StanModels.AbstractMixtureModel

StanModels.AbstractMvNormal

StanModels.AbstractPlot

StanModels.AbstractRGB

StanModels.AbstractSerializer

StanModels.AbstractWeights

StanModels.AlphaColor

StanModels.AnalyticWeights

StanModels.Animation

StanModels.Arcsine

StanModels.AxyY

StanModels.BGR

StanModels.BGRA

StanModels.Bernoulli

StanModels.Beta

StanModels.BetaBinomial

StanModels.BetaPrime

StanModels.BezierCurve

StanModels.Binomial

StanModels.Biweight

StanModels.CIE1931JV_CMF

StanModels.CIE1931J_CMF

StanModels.CIE1931_CMF

StanModels.CIE1964_CMF

StanModels.CMDSTAN_HOME

StanModels.CSV

StanModels.Categorical

StanModels.CategoricalArray

StanModels.CategoricalArrays

StanModels.CategoricalMatrix

StanModels.CategoricalPool

StanModels.CategoricalString

StanModels.CategoricalValue

StanModels.CategoricalVector

StanModels.Cauchy

StanModels.Chains

StanModels.Chi

StanModels.Chisq

StanModels.CmdStan

StanModels.CoefTable

StanModels.Color

StanModels.Color3

StanModels.ColorAlpha

StanModels.ColorGradient

StanModels.ColorTypes

StanModels.Colorant

StanModels.ColorantNormed

StanModels.Colors

StanModels.Continuous

StanModels.ContinuousDistribution

StanModels.ContinuousMatrixDistribution

StanModels.ContinuousMultivariateDistribution

StanModels.ContinuousUnivariateDistribution

StanModels.ConvergenceException

StanModels.Cosine

StanModels.DE_2000

StanModels.DE_94

StanModels.DE_AB

StanModels.DE_BFD

StanModels.DE_CMC

StanModels.DE_DIN99

StanModels.DE_DIN99d

StanModels.DE_DIN99o

StanModels.DE_JPC79

StanModels.DIN99

StanModels.DIN99A

StanModels.DIN99d

StanModels.DIN99dA

StanModels.DIN99o

StanModels.DIN99oA

StanModels.DataFrame

StanModels.DataFrameRow

StanModels.DataFrames

StanModels.DelimitedFiles

StanModels.DiagNormal

StanModels.DiagNormalCanon

StanModels.Diagnose

StanModels.Dirichlet

StanModels.DirichletMultinomial

StanModels.Discrete

StanModels.DiscreteDistribution

StanModels.DiscreteMatrixDistribution

StanModels.DiscreteMultivariateDistribution

StanModels.DiscreteUniform

StanModels.DiscreteUnivariateDistribution

StanModels.Distribution

StanModels.Distributions

StanModels.DoubleExponential

StanModels.ECDF

StanModels.EdgeworthMean

StanModels.EdgeworthSum

StanModels.EdgeworthZ

StanModels.EmpiricalUnivariateDistribution

StanModels.Epanechnikov

StanModels.Erlang

StanModels.Estimator

StanModels.Exponential

StanModels.FDist

StanModels.FisherNoncentralHypergeometric

StanModels.Formatted

StanModels.Fractional

StanModels.Frechet

StanModels.FrequencyWeights

StanModels.FullNormal

StanModels.FullNormalCanon

StanModels.GR

StanModels.Gamma

StanModels.GeneralizedExtremeValue

StanModels.GeneralizedPareto

StanModels.Geometric

StanModels.Gray

StanModels.Gray24

StanModels.GrayA

StanModels.GroupApplied

StanModels.GroupedDataFrame

StanModels.Gumbel

StanModels.HSB

StanModels.HSI

StanModels.HSIA

StanModels.HSL

StanModels.HSLA

StanModels.HSV

StanModels.HSVA

StanModels.Histogram

StanModels.Hypergeometric

StanModels.InverseGamma

StanModels.InverseGaussian

StanModels.InverseWishart

StanModels.IsoNormal

StanModels.IsoNormalCanon

StanModels.JLD

StanModels.KSDist

StanModels.KSOneSided

StanModels.KW

StanModels.Kolmogorov

StanModels.L1dist

StanModels.L2dist

StanModels.LCHab

StanModels.LCHabA

StanModels.LCHuv

StanModels.LCHuvA

StanModels.LMS

StanModels.LMSA

StanModels.Lab

StanModels.LabA

StanModels.Laplace

StanModels.LevelsException

StanModels.Levy

StanModels.Linfdist

StanModels.LocationScale

StanModels.LogNormal

StanModels.Logistic

StanModels.Luv

StanModels.LuvA

StanModels.MCMCChain

StanModels.MLEstimator

StanModels.MSC

StanModels.MatrixDistribution

StanModels.Matrixvariate

StanModels.Missing

StanModels.MissingCategoricalArray

StanModels.MissingCategoricalMatrix

StanModels.MissingCategoricalVector

StanModels.MissingException

StanModels.Missings

StanModels.MixtureModel

StanModels.Multinomial

StanModels.Multivariate

StanModels.MultivariateDistribution

StanModels.MultivariateMixture

StanModels.MultivariateNormal

StanModels.MvLogNormal

StanModels.MvNormal

StanModels.MvNormalCanon

StanModels.MvNormalKnownCov

StanModels.MvTDist

StanModels.NegativeBinomial

StanModels.NonMatrixDistribution

StanModels.NoncentralBeta

StanModels.NoncentralChisq

StanModels.NoncentralF

StanModels.NoncentralHypergeometric

StanModels.NoncentralT

StanModels.Normal

StanModels.NormalCanon

StanModels.NormalInverseGaussian

StanModels.OHLC

StanModels.Optimize

StanModels.OrderedLevelsException

StanModels.Pareto

StanModels.PlotThemes

StanModels.PlotUtils

StanModels.Plots

StanModels.Poisson

StanModels.PoissonBinomial

StanModels.ProbabilityWeights

StanModels.QQPair

StanModels.RDatasets

StanModels.RGB

StanModels.RGB1

StanModels.RGB24

StanModels.RGB4

StanModels.RGBA

StanModels.Rayleigh

StanModels.RealInterval

StanModels.RecipeData

StanModels.RecipesBase

StanModels.RegressionModel

StanModels.Sample

StanModels.Sampleable

StanModels.ScriptEntry

StanModels.Segments

StanModels.Semicircle

StanModels.Serialization

StanModels.Serializer

StanModels.Shape

StanModels.Skellam

StanModels.StanMCMCChain

StanModels.StanModels

StanModels.Stanmodel

StanModels.StatisticalModel

StanModels.StatisticalRethinking

StanModels.Statistics

StanModels.StatsBase

StanModels.StatsFuns

StanModels.StatsPlots

StanModels.SubDataFrame

StanModels.SufficientStats

StanModels.Surface

StanModels.SymTriangularDist

StanModels.TDist

StanModels.Transparent3

StanModels.TransparentColor

StanModels.TransparentGray

StanModels.TransparentRGB

StanModels.TriangularDist

StanModels.Triweight

StanModels.Truncated

StanModels.TruncatedNormal

StanModels.Uniform

StanModels.Univariate

StanModels.UnivariateDistribution

StanModels.UnivariateGMM

StanModels.UnivariateMixture

StanModels.ValueSupport

StanModels.VariateForm

StanModels.Variational

StanModels.VonMises

StanModels.VonMisesFisher

StanModels.WalleniusNoncentralHypergeometric

StanModels.Weibull

StanModels.Weights

StanModels.Wishart

StanModels.XYZ

StanModels.XYZA

StanModels.YCbCr

StanModels.YCbCrA

StanModels.YIQ

StanModels.YIQA

StanModels.ZeroMeanDiagNormal

StanModels.ZeroMeanDiagNormalCanon

StanModels.ZeroMeanFullNormal

StanModels.ZeroMeanFullNormalCanon

StanModels.ZeroMeanIsoNormal

StanModels.ZeroMeanIsoNormalCanon

StanModels.adapted_grid

StanModels.add_backend

StanModels.add_theme

StanModels.addcounts!

StanModels.addrequire

StanModels.adjr2

StanModels.adjr²

StanModels.aggregate

StanModels.aic

StanModels.aicc

StanModels.aliases

StanModels.allowmissing

StanModels.allowmissing!

StanModels.alpha

StanModels.alphacolor

StanModels.andrewsplot

StanModels.andrewsplot!

StanModels.animate

StanModels.annotate!

StanModels.arrow

StanModels.attr!

StanModels.autcorplot

StanModels.autocor

StanModels.autocor!

StanModels.autocorplot

StanModels.autocorplot!

StanModels.autocov

StanModels.autocov!

StanModels.aweights

StanModels.backend

StanModels.backend_name

StanModels.backend_object

StanModels.backends

StanModels.bar

StanModels.bar!

StanModels.barh

StanModels.barh!

StanModels.barhist

StanModels.barhist!

StanModels.base_color_type

StanModels.base_colorant_type

StanModels.bbox

StanModels.betaccdf

StanModels.betacdf

StanModels.betainvccdf

StanModels.betainvcdf

StanModels.betainvlogccdf

StanModels.betainvlogcdf

StanModels.betalogccdf

StanModels.betalogcdf

StanModels.betalogpdf

StanModels.betapdf

StanModels.bic

StanModels.binaryentropy

StanModels.binomccdf

StanModels.binomcdf

StanModels.binominvccdf

StanModels.binominvcdf

StanModels.binominvlogccdf

StanModels.binominvlogcdf

StanModels.binomlogccdf

StanModels.binomlogcdf

StanModels.binomlogpdf

StanModels.binompdf

StanModels.blue

StanModels.boxplot

StanModels.boxplot!

StanModels.brush

StanModels.by

StanModels.canonform

StanModels.categorical

StanModels.categorical!

StanModels.ccdf

StanModels.ccolor

StanModels.cdensity

StanModels.cdensity!

StanModels.cdf

StanModels.center

StanModels.cf

StanModels.cgf

StanModels.cgrad

StanModels.cgradients

StanModels.chisqccdf

StanModels.chisqcdf

StanModels.chisqinvccdf

StanModels.chisqinvcdf

StanModels.chisqinvlogccdf

StanModels.chisqinvlogcdf

StanModels.chisqlogccdf

StanModels.chisqlogcdf

StanModels.chisqlogpdf

StanModels.chisqpdf

StanModels.cie_color_match

StanModels.circvar

StanModels.clibraries

StanModels.clibrary

StanModels.closeall

StanModels.coalesce

StanModels.coef

StanModels.coefnames

StanModels.coeftable

StanModels.color

StanModels.color_type

StanModels.coloralpha

StanModels.colordiff

StanModels.colormap

StanModels.colormatch

StanModels.colwise

StanModels.combine

StanModels.comp1

StanModels.comp2

StanModels.comp3

StanModels.compact

StanModels.competerank

StanModels.complete_cases

StanModels.complete_cases!

StanModels.completecases

StanModels.component

StanModels.components

StanModels.componentwise_logpdf

StanModels.componentwise_pdf

StanModels.compress

StanModels.concentration

StanModels.confint

StanModels.contour

StanModels.contour!

StanModels.contour3d

StanModels.contour3d!

StanModels.contourf

StanModels.contourf!

StanModels.convert

StanModels.convert_a3d

StanModels.coords

StanModels.cor

StanModels.cor2cov

StanModels.corkendall

StanModels.corner

StanModels.corner!

StanModels.cornerplot

StanModels.cornerplot!

StanModels.corrplot

StanModels.corrplot!

StanModels.corspearman

StanModels.counteq

StanModels.countmap

StanModels.countne

StanModels.counts

StanModels.cov

StanModels.cov2cor

StanModels.cquantile

StanModels.creator

StanModels.crosscor

StanModels.crosscor!

StanModels.crosscov

StanModels.crosscov!

StanModels.crossentropy

StanModels.cumulant

StanModels.current

StanModels.curve_points

StanModels.curves

StanModels.curves!

StanModels.cut

StanModels.cvec

StanModels.dataset

StanModels.dataviewer

StanModels.decompress

StanModels.default

StanModels.default_cgrad

StanModels.delete!

StanModels.deletecols!

StanModels.deleterows!

StanModels.denserank

StanModels.density

StanModels.density!

StanModels.densityplot

StanModels.describe

StanModels.deserialize

StanModels.deuteranopic

StanModels.deviance

StanModels.dim

StanModels.disallowmissing

StanModels.disallowmissing!

StanModels.discretediag

StanModels.distinguishable_colors

StanModels.diverging_palette

StanModels.dof

StanModels.dof_residual

StanModels.droplevels!

StanModels.dropmissing

StanModels.dropmissing!

StanModels.ea_histogram

StanModels.ea_histogram!

StanModels.ecdf

StanModels.eltypes

StanModels.entropy

StanModels.estimate

StanModels.eval

StanModels.expected_logdet

StanModels.failprob

StanModels.fdistccdf

StanModels.fdistcdf

StanModels.fdistinvccdf

StanModels.fdistinvcdf

StanModels.fdistinvlogccdf

StanModels.fdistinvlogcdf

StanModels.fdistlogccdf

StanModels.fdistlogcdf

StanModels.fdistlogpdf

StanModels.fdistpdf

StanModels.findat

StanModels.fit

StanModels.fit!

StanModels.fit_map

StanModels.fit_map!

StanModels.fit_mle

StanModels.fit_mle!

StanModels.fitted

StanModels.font

StanModels.fourinvπ

StanModels.fourπ

StanModels.frame

StanModels.freecumulant

StanModels.fweights

StanModels.gammaccdf

StanModels.gammacdf

StanModels.gammainvccdf

StanModels.gammainvcdf

StanModels.gammainvlogccdf

StanModels.gammainvlogcdf

StanModels.gammalogccdf

StanModels.gammalogcdf

StanModels.gammalogpdf

StanModels.gammapdf

StanModels.gamutmax

StanModels.gamutmin

StanModels.gelmandiag

StanModels.generate

StanModels.generate_s — Function.

generate_s

Utility function to generate all notebooks and chapters from scripts in the scripts directory.

Method

generate_s(sd = script_dict_s)

Required arguments

None, all notebooks/.. and chapters/.. files are regenerated.

generate_s

Generate notebooks and scripts in a single chapter.

Method

generate_s(chapter::AbstractString)

Required arguments

Generate notebooks and scripts in a single chapter, e.g. generate("04")

generate_s

Generate a single notebook and script

Method

generate_s(chapter::AbstractString, file::AbstractString)

Required arguments

Generate notebook and script file in chapter, e.g. generate("04", "m4.1d.jl") or generate("04/m4.1d.jl")

StanModels.genmean

StanModels.genvar

StanModels.geomean

StanModels.get_color_palette

StanModels.getindex

StanModels.gewekediag

StanModels.gif

StanModels.gkldiv

StanModels.gr

StanModels.gradlogpdf

StanModels.gray

StanModels.green

StanModels.grid

StanModels.groupby

StanModels.groupedbar

StanModels.groupedbar!

StanModels.groupindices

StanModels.groupvars

StanModels.gui

StanModels.halfπ

StanModels.harmmean

StanModels.hasfinitesupport

StanModels.hdf5

StanModels.head

StanModels.heatmap

StanModels.heatmap!

StanModels.heideldiag

StanModels.hex

StanModels.hexbin

StanModels.hexbin!

StanModels.hist

StanModels.histogram

StanModels.histogram!

StanModels.histogram2d

StanModels.histogram2d!

StanModels.histogramplot

StanModels.hline

StanModels.hline!

StanModels.hspan

StanModels.hspan!

StanModels.hyperccdf

StanModels.hypercdf

StanModels.hyperinvccdf

StanModels.hyperinvcdf

StanModels.hyperinvlogccdf

StanModels.hyperinvlogcdf

StanModels.hyperlogccdf

StanModels.hyperlogcdf

StanModels.hyperlogpdf

StanModels.hyperpdf

StanModels.include

StanModels.indexmap

StanModels.indicatormat

StanModels.informationmatrix

StanModels.inline

StanModels.insert!

StanModels.insertcols!

StanModels.inspectdr

StanModels.insupport

StanModels.inv2π

StanModels.inv4π

StanModels.invcov

StanModels.inverse_rle

StanModels.invisible

StanModels.invlogccdf

StanModels.invlogcdf

StanModels.invscale

StanModels.invsoftplus

StanModels.invsqrt2

StanModels.invsqrt2π

StanModels.invπ

StanModels.iqr

StanModels.isbounded

StanModels.isdark

StanModels.isfitted

StanModels.isleptokurtic

StanModels.islinear

StanModels.islowerbounded

StanModels.ismesokurtic

StanModels.ismissing

StanModels.ismmappable

StanModels.isordered

StanModels.isplatykurtic

StanModels.isprobvec

StanModels.isupperbounded

StanModels.iter_segments

StanModels.jldopen

StanModels.kde

StanModels.keys

StanModels.kldivergence

StanModels.kurtosis

StanModels.length

StanModels.levels

StanModels.levels!

StanModels.levelsmap

StanModels.leverage

StanModels.link

StanModels.load

StanModels.location

StanModels.location!

StanModels.log1mexp

StanModels.log1pexp

StanModels.log1pmx

StanModels.log1psq

StanModels.log2mexp

StanModels.log2π

StanModels.log4π

StanModels.logaddexp

StanModels.logccdf

StanModels.logcdf

StanModels.logdetcov

StanModels.logexpm1

StanModels.loghalf

StanModels.logistic

StanModels.logit

StanModels.loglikelihood

StanModels.logmvgamma

StanModels.logmxp1

StanModels.logpdf

StanModels.logpdf!

StanModels.logsumexp

StanModels.logtwo

StanModels.logπ

StanModels.lstirling_asym

StanModels.mad

StanModels.mad!

StanModels.map

StanModels.mapc

StanModels.mapcols

StanModels.mapreducec

StanModels.marginalhist

StanModels.marginalhist!

StanModels.maxad

StanModels.mean

StanModels.mean!

StanModels.mean_and_cov

StanModels.mean_and_std

StanModels.mean_and_var

StanModels.meanad

StanModels.meandir

StanModels.meanform

StanModels.meanlogx

StanModels.meanplot

StanModels.meanplot!

StanModels.meanresponse

StanModels.median

StanModels.median!

StanModels.melt

StanModels.meltdf

StanModels.merge!

StanModels.mgf

StanModels.middle

StanModels.midpoints

StanModels.missing

StanModels.missings

StanModels.mixeddensity

StanModels.mixeddensity!

StanModels.mixeddensityplot

StanModels.mode

StanModels.model_response

StanModels.modelmatrix

StanModels.modes

StanModels.moment

StanModels.mov

StanModels.mp4

StanModels.msd

StanModels.mss

StanModels.names!

StanModels.nbetaccdf

StanModels.nbetacdf

StanModels.nbetainvccdf

StanModels.nbetainvcdf

StanModels.nbetainvlogccdf

StanModels.nbetainvlogcdf

StanModels.nbetalogccdf

StanModels.nbetalogcdf

StanModels.nbetalogpdf

StanModels.nbetapdf

StanModels.nbinomccdf

StanModels.nbinomcdf

StanModels.nbinominvccdf

StanModels.nbinominvcdf

StanModels.nbinominvlogccdf

StanModels.nbinominvlogcdf

StanModels.nbinomlogccdf

StanModels.nbinomlogcdf

StanModels.nbinomlogpdf

StanModels.nbinompdf

StanModels.ncategories

StanModels.nchisqccdf

StanModels.nchisqcdf

StanModels.nchisqinvccdf

StanModels.nchisqinvcdf

StanModels.nchisqinvlogccdf

StanModels.nchisqinvlogcdf

StanModels.nchisqlogccdf

StanModels.nchisqlogcdf

StanModels.nchisqlogpdf

StanModels.nchisqpdf

StanModels.ncol

StanModels.ncomponents

StanModels.nfdistccdf

StanModels.nfdistcdf

StanModels.nfdistinvccdf

StanModels.nfdistinvcdf

StanModels.nfdistinvlogccdf

StanModels.nfdistinvlogcdf

StanModels.nfdistlogccdf

StanModels.nfdistlogcdf

StanModels.nfdistlogpdf

StanModels.nfdistpdf

StanModels.nobs

StanModels.nonunique

StanModels.norepeats

StanModels.normccdf

StanModels.normcdf

StanModels.norminvccdf

StanModels.norminvcdf

StanModels.norminvlogccdf

StanModels.norminvlogcdf

StanModels.normlogccdf

StanModels.normlogcdf

StanModels.normlogpdf

StanModels.normpdf

StanModels.nquantile

StanModels.nrow

StanModels.nsamples

StanModels.ntdistccdf

StanModels.ntdistcdf

StanModels.ntdistinvccdf

StanModels.ntdistinvcdf

StanModels.ntdistinvlogccdf

StanModels.ntdistinvlogcdf

StanModels.ntdistlogccdf

StanModels.ntdistlogcdf

StanModels.ntdistlogpdf

StanModels.ntdistpdf

StanModels.ntrials

StanModels.nullable!

StanModels.nulldeviance

StanModels.nullloglikelihood

StanModels.o_delete

StanModels.ohlc

StanModels.ohlc!

StanModels.optimize_datetime_ticks

StanModels.optimize_ticks

StanModels.order

StanModels.ordered

StanModels.ordered!

StanModels.ordinalrank

StanModels.pacf

StanModels.pacf!

StanModels.palette

StanModels.params

StanModels.params!

StanModels.partialcor

StanModels.partype

StanModels.passmissing

StanModels.path3d

StanModels.path3d!

StanModels.pdf

StanModels.percentile

StanModels.permutecols!

StanModels.pgfplots

StanModels.pie

StanModels.pie!

StanModels.plot

StanModels.plot!

StanModels.plot3d

StanModels.plot3d!

StanModels.plot_color

StanModels.plotarea

StanModels.plotattr

StanModels.plotly

StanModels.plotlyjs

StanModels.plots_heatmap

StanModels.plots_heatmap!

StanModels.png

StanModels.poisccdf

StanModels.poiscdf

StanModels.poisinvccdf

StanModels.poisinvcdf

StanModels.poisinvlogccdf

StanModels.poisinvlogcdf

StanModels.poislogccdf

StanModels.poislogcdf

StanModels.poislogpdf

StanModels.poispdf

StanModels.pool

StanModels.pool!

StanModels.portfoliocomposition

StanModels.portfoliocomposition!

StanModels.predict

StanModels.predict!

StanModels.probs

StanModels.probval

StanModels.proportionmap

StanModels.proportions

StanModels.protanopic

StanModels.psnr

StanModels.pweights

StanModels.pyplot

StanModels.qqbuild

StanModels.qqnorm

StanModels.qqnorm!

StanModels.qqplot

StanModels.qqplot!

StanModels.quantile

StanModels.quantile!

StanModels.quartπ

StanModels.quiver

StanModels.quiver!

StanModels.r2

StanModels.rafterydiag

StanModels.rate

StanModels.readdlm

StanModels.readmmap

StanModels.readtable

StanModels.recode

StanModels.recode!

StanModels.red

StanModels.reducec

StanModels.rel_path

StanModels.rel_path_s — Function.

relpaths

Relative path using the StanModels src/ directory.

Example to get access to the chapters subdirectories

rel_path_s("..", "chapters")

StanModels.rename

StanModels.rename!

StanModels.renyientropy

StanModels.residuals

StanModels.response

StanModels.rgb_string

StanModels.rgba_string

StanModels.rle

StanModels.rmsd

StanModels.rotate

StanModels.rotate!

StanModels.rss

StanModels.r²

StanModels.sample

StanModels.sample!

StanModels.samplepair

StanModels.sampler

StanModels.save

StanModels.savefig

StanModels.scale

StanModels.scale!

StanModels.scatter

StanModels.scatter!

StanModels.scatter3d

StanModels.scatter3d!

StanModels.scatterhist

StanModels.scatterhist!

StanModels.scattermat

StanModels.score

StanModels.script_dict

StanModels.script_dict_s

StanModels.scriptentry

StanModels.sem

StanModels.sequential_palette

StanModels.serialize

StanModels.set_cmdstan_home!

StanModels.set_theme

StanModels.setindex!

StanModels.shape

StanModels.shape_coords

StanModels.show

StanModels.showall

StanModels.showcols

StanModels.showgradient

StanModels.showgradient!

StanModels.showlibrary

StanModels.showlibrary!

StanModels.showtheme

StanModels.showtheme!

StanModels.skewness

StanModels.softmax

StanModels.softmax!

StanModels.softplus

StanModels.span

StanModels.spy

StanModels.spy!

StanModels.sqL2dist

StanModels.sqmahal

StanModels.sqmahal!

StanModels.sqrt2

StanModels.sqrt2π

StanModels.sqrt3

StanModels.sqrt4π

StanModels.sqrthalfπ

StanModels.sqrtπ

StanModels.src_path_s

StanModels.srdistccdf

StanModels.srdistcdf

StanModels.srdistinvccdf

StanModels.srdistinvcdf

StanModels.srdistinvlogccdf

StanModels.srdistinvlogcdf

StanModels.srdistlogccdf

StanModels.srdistlogcdf

StanModels.stack

StanModels.stackdf

StanModels.stan

StanModels.std

StanModels.stderror

StanModels.stdlogx

StanModels.stdm

StanModels.stephist

StanModels.stephist!

StanModels.sticks

StanModels.sticks!

StanModels.stroke

StanModels.sub

StanModels.succprob

StanModels.suffstats

StanModels.summarystats

StanModels.support

StanModels.surface

StanModels.surface!

StanModels.tail

StanModels.tdistccdf

StanModels.tdistcdf

StanModels.tdistinvccdf

StanModels.tdistinvcdf

StanModels.tdistinvlogccdf

StanModels.tdistinvlogcdf

StanModels.tdistlogccdf

StanModels.tdistlogcdf

StanModels.tdistlogpdf

StanModels.tdistpdf

StanModels.test_distr

StanModels.test_examples

StanModels.test_samples

StanModels.testhist

StanModels.testhist!

StanModels.text

StanModels.theme

StanModels.tiedrank

StanModels.title!

StanModels.totalvar

StanModels.traceplot

StanModels.traceplot!

StanModels.transform

StanModels.translate

StanModels.translate!

StanModels.trim

StanModels.trim!

StanModels.trimvar

StanModels.tritanopic

StanModels.truncate_module_path

StanModels.twinx

StanModels.twoinvπ

StanModels.twoπ

StanModels.uncompact

StanModels.unicodeplots

StanModels.unique!

StanModels.unstack

StanModels.values

StanModels.var

StanModels.variation

StanModels.varlogx

StanModels.varm

StanModels.vcat

StanModels.vcov

StanModels.violin

StanModels.violin!

StanModels.vline

StanModels.vline!

StanModels.vspan

StanModels.vspan!

StanModels.weighted_color_mean

StanModels.weights

StanModels.whitebalance

StanModels.winsor

StanModels.winsor!

StanModels.wireframe

StanModels.wireframe!

StanModels.with

StanModels.wmean

StanModels.wmean!

StanModels.wmedian

StanModels.wquantile

StanModels.wrap

StanModels.writedlm

StanModels.writetable

StanModels.wsample

StanModels.wsample!

StanModels.wsum

StanModels.wsum!

StanModels.xaxis!

StanModels.xflip!

StanModels.xgrid!

StanModels.xlabel!

StanModels.xlims

StanModels.xlims!

StanModels.xlogx

StanModels.xlogy

StanModels.xticks!

StanModels.xyY

StanModels.xyYA

StanModels.yaxis!

StanModels.yflip!

StanModels.ygrid!

StanModels.ylabel!

StanModels.ylims

StanModels.ylims!

StanModels.yticks!

StanModels.zlims

StanModels.zlims!

StanModels.zscore

StanModels.zscore!