Docstrings

ComputedFieldTypes.@computed — Macro.

@computed type

Allows declaration of a type expression where some of the fields types are arbitrary computed functions of other type parameters. It's suggested that those expressions be const (return the same value for the same inputs), but that is not essential.

ComputedFieldTypes.ComputedFieldTypes

ComputedFieldTypes._computed — Function.

_computed(typeexpr::Expr)

the bulk of the work to compute the AST transform

ComputedFieldTypes.eval

ComputedFieldTypes.fulltype — Function.

fulltype(::Type)

Fill in the computed type parameters for T (if applicable and possible)

ComputedFieldTypes.getenv! — Function.

getenv!(expr, tvars, defs)

replace anything that isn't computable by apply_type with a dummy type-variable

ComputedFieldTypes.include

ComputedFieldTypes.make_Type_expr — Function.

make_Type_expr(tname, decl_tvars)

make the ::Type{T} expression that is equivalent to the original type declaration

ComputedFieldTypes.make_curly — Function.

make_curly(expr)

given an apply-type expression (T or T{...}), return Expr(:curly, T, ...)

ComputedFieldTypes.make_fulltype_expr — Function.

make_fulltype_expr(tname, decl_tvars, def)

compute the leaf T{def} expression that is equivalent for the new type declaration

ComputedFieldTypes.make_new_expr — Function.

make_new_expr(tname, decl_tvars, def)

compute the leaf T{...} expression that describes the new type declaration

ComputedFieldTypes.rewrite_new! — Function.

rewrite_new!(expr, tname::Symbol, decl_tvars, def)

rewrite the constructors to capture only the intended values