Docstrings

BayesOpt.BayesOpt — Module.

Example

using BayesOpt
config = ConfigParameters()         # calls initialize_parameters_to_default of the C API
set_kernel!(config, "kMaternARD5")  # calls set_kernel of the C API
config.sc_type = SC_MAP
f(x) = sum(x .^ 2)
lowerbound = [-2., -2.]; upperbound = [2., 2.]
optimizer, optimum = bayes_optimization(f, lowerbound, upperbound, config)

Exports: KernelParameters, MeanParameters, ConfigParameters, bayes_optimization, bayes_optimization_disc, bayes_optimization_categorical, set_kernel!, set_mean!, set_criteria!, set_surrogate!, set_log_file!, set_load_file!, set_save_file!, set_learning!, set_score!, L_FIXED, L_EMPIRICAL, L_DISCRETE, L_MCMC, L_ERROR, SC_MTL, SC_ML, SC_MAP, SC_LOOCV, SC_ERROR

See also https://rmcantin.bitbucket.io/html/usemanual.html

BayesOpt.ConfigParameters — Type.

ConfigParameters()

Returns default parameters of BayesOpt (see initializeparametersto_default in the C API).

BayesOpt.KernelParameters

BayesOpt.L_DISCRETE

BayesOpt.L_EMPIRICAL

BayesOpt.L_ERROR

BayesOpt.L_FIXED

BayesOpt.L_MCMC

BayesOpt.LearningType

BayesOpt.MeanParameters

BayesOpt.SC_ERROR

BayesOpt.SC_LOOCV

BayesOpt.SC_MAP

BayesOpt.SC_ML

BayesOpt.SC_MTL

BayesOpt.ScoreType

BayesOpt.__init__

BayesOpt.bayes_optimization — Function.

bayes_optimization(func, lb, ub, config)

Runs continuous Bayesian optimization on func that takes vectors of length d as argument and returns a real number, within the box defined by the lowerbounds lb (a vector of length d with lowerbounds for each dimension) and upperbounds ub, using config (see ConfigParameters).

BayesOpt.bayes_optimization_categorical — Function.

bayes_optimization_categorical(func, categories, config)

Runs Bayesian optimization on func that takes vectors of length d as argument and returns a real number, with categories array of size d with the number of categories per dimension, using config (see ConfigParameters).

BayesOpt.bayes_optimization_disc — Function.

bayes_optimization_disc(func, valid_x, config)

Runs Bayesian optimization on func that takes vectors of length d as argument and returns a real number, on valid_x an array of valid points (vectors of length d) using config (see ConfigParameters).

BayesOpt.check_deps

BayesOpt.depsjl_path

BayesOpt.eval

BayesOpt.include

BayesOpt.libbayesopt

BayesOpt.libnlopt

BayesOpt.prepare_cargs

BayesOpt.set_criteria! — Function.

set_criteria!(config, name) set criteria in config to name.

BayesOpt.set_kernel! — Function.

set_kernel!(config, name) set kernel in config to name.

BayesOpt.set_learning! — Function.

set_learning!(config, name) set learning in config to name.

BayesOpt.set_load_file! — Function.

set_load_file!(config, name) set load file in config to name.

BayesOpt.set_log_file! — Function.

set_log_file!(config, name) set log file in config to name.

BayesOpt.set_mean! — Function.

set_mean!(config, name) set mean in config to name.

BayesOpt.set_save_file! — Function.

set_save_file!(config, name) set save file in config to name.

BayesOpt.set_score! — Function.

set_score!(config, name) set score in config to name.

BayesOpt.set_surrogate! — Function.

set_surrogate!(config, name) set surrogate in config to name.