Docstrings

RiemannTheta.RiemannTheta

RiemannTheta.deriv_prod — Function.

  deriv_prod(n::Vector{Float64},
             intshift::Vector{Float64},
             derivs::Vector{Vector{ComplexF64}})::ComplexF64

Compute the real and imaginary parts of the product ___ | | 2π * I <d, n-intshift> | | d in derivs for a given n in ZZ^g.

Parameters

n : An integer vector in the finite sum ellipsoid.

intshift : The integer part of Yinv*y.

derivreal, derivimag : The real and imaginary parts of the derivative directional vectors.

Returns

The complex "derivative product".

RiemannTheta.eval

RiemannTheta.exponential_part — Function.

     exponential_part(zs::Vector{Vector{ComplexF64}},
                      Ω::Matrix{ComplexF64})::Vector{Float64}

Return the value of the exponential part of the Riemann theta function for Ω and all z in zs.

Parameters

zs : A vector of complex vectors at which to evaluate the Riemann theta function.

Omega : A Riemann matrix.

Returns

The value of the exponential part of the Riemann theta function at each point appearing in zs.

RiemannTheta.exppart

RiemannTheta.finite_sum — Function.

function finite_sum(X::Matrix{Float64},
                    Yinv::Matrix{Float64},
                    T::Matrix{Float64},
                    z::Vector{Vector{ComplexF64}},
                    S::Vector{Vector{Float64}},
                    derivs::Vector{Vector{ComplexF64}})

Computes the real and imaginary parts of the finite sum with derivatives.

Parameters

X, Yinv, T : Row-major matrices such that the Riemann matrix, Omega is equal to (X + iY). T is the Cholesky decomposition of Y. z : Input vectors. S : The set of points in ZZ^g over which to compute the finite sum derivs : the derivative directional vectors.

Returns

The finite sums for each z.

RiemannTheta.finite_sum_large

RiemannTheta.finite_sum_small

RiemannTheta.include

RiemannTheta.innerpoints

RiemannTheta.lll — Function.

lll(b₀::AbstractMatrix{Float64})::AbstractMatrix{Float64}

Performs Lenstra-Lenstra-Lovasv reduction on the n vectors in b₀.

Parameters

b₀ : vector of n vectors of n Float64

Returns

The LLL reduction of the vectors in b₀

RiemannTheta.normpart

RiemannTheta.oscillatory_part — Function.

     oscillatory_part(zs::Vector{Vector{ComplexF64}},
                      Ω::Matrix{ComplexF64};
                      eps::Float64=1e-8,
                      derivs::Vector{Vector{ComplexF64}}=Vector{ComplexF64}[],
                      accuracy_radius::Float64=5.)::Vector{ComplexF64}

Return the value of the oscillatory part of the Riemann theta function for Ω and all z in zs if derivs is empty, or the derivatives at all z in zs for the given directional derivatives in derivs.

Parameters

zs : A vector of complex vectors at which to evaluate the Riemann theta function.

Omega : A Riemann matrix.

eps : (Default: 1e-8) The desired numerical accuracy.

derivs : A vector of complex vectors giving a directional derivative.

accuracy_radius : (Default: 5.) The radius from the g-dimensional origin

where the requested accuracy of the Riemann theta is guaranteed when computing derivatives. Not used if no derivatives of theta are requested.

Returns

The value of the oscillatory part of the Riemann theta function at each point appearing in z.

RiemannTheta.padding

RiemannTheta.radius — Function.

radius(ϵ::Float64,
       T::Matrix{Float64},
       derivs::Vector{Vector{Float64}} = Vector{Float64}[],
       accuracy_radius::Float64 = 5.)

Returns the primary radius of the bounding ellipsoid for computing the Riemann theta function up to accuracy ϵ.

The derivative oscillatory part of the Riemann theta function has linear growth in :math:z along the directions of the columns of the Riemann matrix. accuracy_radius is used to determine a radius resulting in an accurate Riemann theta for all

.. math ::

||z|| < 	ext{accuracy_radius}.

Parameters

ϵ : Requested accuracy.

T : A gxg matrix representing the Cholesky decomposition of the imaginary part of a Riemann matrix.

derivs : (Default: []) A list of directional derivatives. The number of directional derivatives is the order, N, of the derivative we wish to compute.

accuracy_radius : (Default: 5) Radius for guaranteed region of accuracy. See above.

Returns

radius : The initial radius of the bounding ellipsoid used to truncate the Riemann theta function to desired accuracy.

RiemannTheta.radius0 — Function.

radius0(eps::Float64, r::Float64, g::Int)::Float64

Compute the radius with no derivatives.

RiemannTheta.radiusN — Function.

radiusN(eps, r, T, derivs, accuracy_radius=5)::Float64

Compute the radius with N derivatives.

Parameters

eps : Requested accuracy.

r : The length of the shortest lattice vector in the LLL reduction of the Cholesky decomposition of the imaginary part of the Riemann matrix.

T : A gxg matrix representing the Cholesky decomposition of the imaginary part of a Riemann matrix.

derivs : A list of directional derivatives. The number of directional derivatives is the order, N, of the derivative we wish to compute.

accuracy_radius : Radius for guaranteed region of accuracy. See :func:radius.

Returns

radius : The initial radius of the bounding ellipsoid used to truncate the Riemann theta function to desired accuracy.

RiemannTheta.riemanntheta — Function.

 riemanntheta(zs::Vector{Vector{ComplexF64}},

Ω::Matrix{ComplexF64}; eps::Float64=1e-8, derivs::Vector{Vector{ComplexF64}}=Vector{ComplexF64}[], accuracy_radius::Float64=5.)::Vector{ComplexF64}

Return the value of the Riemann theta function for Ω and all z in zs if derivs is empty, or the derivatives at all z in zs for the given directional derivatives in derivs.

Parameters

zs : A vector of complex vectors at which to evaluate the Riemann theta function.

Omega : A Riemann matrix.

eps : (Default: 1e-8) The desired numerical accuracy.

derivs : A vector of complex vectors giving a directional derivative.

accuracy_radius : (Default: 5.) The radius from the g-dimensional origin

where the requested accuracy of the Riemann theta is guaranteed when computing derivatives. Not used if no derivatives of theta are requested.

Returns

The value (or derivative) of the Riemann theta function at each point in zs.