Docstrings

UnitfulUS.@us_str — Macro.

macro us_str(unit)

String macro to easily recall U.S. customary units located in the UnitfulUS package. Although all unit symbols in that package are suffixed with _us, the suffix should not be used when using this macro.

Note that what goes inside must be parsable as a valid Julia expression.

Examples:

julia> 1.0us"tbsp"
1.0 m s^-1

julia> 1.0us"syd" - 1.0u"yd"
1.0 m N

UnitfulUS.UnitfulUS

UnitfulUS.__init__

UnitfulUS.allowed_funcs

UnitfulUS.bushel_us

UnitfulUS.cup_us

UnitfulUS.cwt_us

UnitfulUS.dottify

UnitfulUS.drypt_us

UnitfulUS.dryqt_us

UnitfulUS.eval

UnitfulUS.fldr_us

UnitfulUS.floz_us

UnitfulUS.gal_us

UnitfulUS.gill_us

UnitfulUS.include

UnitfulUS.localunits

UnitfulUS.minim_us

UnitfulUS.pk_us

UnitfulUS.pt_us

UnitfulUS.qt_us

UnitfulUS.replace_value

UnitfulUS.sac_us

UnitfulUS.sch_us

UnitfulUS.sft_us

UnitfulUS.sfur_us

UnitfulUS.sinch_us

UnitfulUS.slea_us

UnitfulUS.sli_us

UnitfulUS.smi_us

UnitfulUS.srd_us

UnitfulUS.syd_us

UnitfulUS.tbsp_us

UnitfulUS.ton_us

UnitfulUS.town_us

UnitfulUS.tsp_us

UnitfulUS.ustrcheck_bool