Docstrings

LazyJSON.!=

LazyJSON.==

LazyJSON.Array

LazyJSON.Collection

LazyJSON.IOStrings

LazyJSON.Indexes — Type.

Iterate over the byte-indexes of the values in a JSON Array, or the byte-indexes of the alternating field names and values in a JSON Object.

LazyJSON.JSON

LazyJSON.JSONjl

LazyJSON.LazyJSON

LazyJSON.Number

LazyJSON.Object

LazyJSON.ParseError

LazyJSON.PropertyDicts

LazyJSON.SplicedStrings

LazyJSON.String — Type.

Values are represented by a reference to the JSON text String and the byte index of the value text.

String: {"foo": 1, "bar": [1, 2, 3, "four"]}
        ▲                 ▲      ▲  ▲
        │                 │      │  │
        ├──────────────┐  │      │  │
        │   JSON.Array(s, i=9)   │  │   == Any[1, 2, 3, "four"]
        │                        │  │
        ├───────────────┐  ┌─────┘  │
        │   JSON.Number(s, i=16)    │   == 3
        │                           │
        ├───────────────┐  ┌────────┘
        │   JSON.String(s, i=19)        == "four"
        │
        └───────────────┬──┐
            JSON.Object(s, i=1)

LazyJSON.SupportedString

LazyJSON.Unsafe

LazyJSON.Value

LazyJSON._get

LazyJSON._iterate

LazyJSON.collection_length — Function.

Count the number of values in an Array or Object. For an Object, the count includes the field names.

LazyJSON.decimal

LazyJSON.enable_getproperty

LazyJSON.eval

LazyJSON.flat_array

LazyJSON.flat_object

LazyJSON.flatten

LazyJSON.get_field — Function.

Get a field from a JSON.Object starting at start_i. Returns start_i for next field and field value.

LazyJSON.get_ic — Function.

Get the index i and first byte c of the field value for key in an Object.

LazyJSON.getc — Function.

Get byte i in string s without bounds checking.

LazyJSON.getflat — Function.

Get a flattened (non-lazy) Julia object for a value in a JSON text; and the index of the last character of the value.

LazyJSON.getindex_ic — Function.

Get the index i and first byte c of the nth value in an Array.

LazyJSON.getpath — Function.

Find the value at a specified key path in a JSON text.

s, a JSON text

path, A vector of Ints and Strings representing JSON Array indexes and JSON Object field names.

i, start index

c, byte at index i

LazyJSON.getvalue — Function.

Get a JSON value object for a value in a JSON text.

s, the JSON text.

i, byte index of the value in JSON text.

c, first byte of the value.

LazyJSON.hexvalue — Function.

Lookup tables for hex digits.

LazyJSON.include

LazyJSON.indexes

LazyJSON.isbegin — Function.

https://tools.ietf.org/html/rfc7159#section-2

These are the six structural characters:
    begin-array     = ws %x5B ws  ; [ left square bracket
    begin-object    = ws %x7B ws  ; { left curly bracket
    end-array       = ws %x5D ws  ; ] right square bracket
    end-object      = ws %x7D ws  ; } right curly bracket
    name-separator  = ws %x3A ws  ; : colon
    value-separator = ws %x2C ws  ; , comma

LazyJSON.isend

LazyJSON.isescape — Function.

Is i the index of the start of an escape sequence in a JSON string s? Yes if s[i] == '\', unless the escape is escaped itself.

LazyJSON.ishexdigit

LazyJSON.isnoise — Function.

The tokeniser ignores whitespace and separators between values.

LazyJSON.isnull — Function.

jl_alloc_string allocates n + 1 bytes and sets the last byte to 0x00 https://github.com/JuliaLang/julia/blob/master/src/array.c#L464

LazyJSON.isnum — Function.

First character of a Start of a JSON Number https://tools.ietf.org/html/rfc7159#section-6

LazyJSON.isreserved

LazyJSON.iswhitespace — Function.

https://tools.ietf.org/html/rfc7159#section-2

Insignificant whitespace is allowed before or after any of the six
structural characters.

ws = *(
        %x20 /              ; Space
        %x09 /              ; Horizontal tab
        %x0A /              ; Line feed or New line
        %x0D )              ; Carriage return

LazyJSON.json_char — Function.

Return a Char from a JSON string s at index i. Unescpe character if needed.

LazyJSON.json_unescape_char — Function.

Return the character code for an escaped character in a JSON string s starting at index i (the \ character).

LazyJSON.jsonstring — Function.

JSON text representation of x

LazyJSON.lastindex_of_collection — Function.

Find index of last byte of a JSON Collection.

LazyJSON.lastindex_of_number — Function.

Find index of last byte of a JSON Number.

LazyJSON.lastindex_of_string — Function.

Find index of last byte of a JSON String.

LazyJSON.lastindex_of_token — Function.

Find last index of a value or collection begin/end token end in a JSON text.

s, a JSON text

i, token start index

c, byte at index i

LazyJSON.lastindex_of_value — Function.

Find index of last byte of a JSON value.

s, a JSON text

i, value start index

c, byte at index i

LazyJSON.memcmp

LazyJSON.next_c

LazyJSON.next_i

LazyJSON.next_ic — Function.

Increment i and get byte i in string s without bounds checking.

LazyJSON.nextindex

LazyJSON.old_parse_number

LazyJSON.parse — Function.

See LazyJSON.value(jsontext) -> JSON.Value.

LazyJSON.parse_number — Function.

https://tools.ietf.org/html/rfc7159#section-6

  number = [ minus ] int [ frac ] [ exp ]

  decimal-point = %x2E       ; .

  digit1-9 = %x31-39         ; 1-9

  e = %x65 / %x45            ; e E

  exp = e [ minus / plus ] 1*DIGIT

  frac = decimal-point 1*DIGIT

  int = zero / ( digit1-9 *DIGIT )

  minus = %x2D               ; -

  plus = %x2B                ; +

  zero = %x30                ; 0

LazyJSON.parse_string

LazyJSON.promotejson — Function.

Promote a JSON.Vaue to an equivalent Base type.

LazyJSON.scan_string — Function.

Find index of last byte of a JSON String and check for existence of the escape character.

https://tools.ietf.org/html/rfc7159#section-7

string = quotation-mark *char quotation-mark
char = unescaped / escape ( ... )

LazyJSON.setc_utf8 — Function.

Write a Unicode chatacter c into a String s as UTF8 at byte index i.

LazyJSON.skip_noise — Function.

Skip over non-value characters in String s starting at index i + 1.

LazyJSON.skip_whitespace — Function.

Skip over whitespace in String s starting at index i.

LazyJSON.splice — Function.

Replace value v with x.

LazyJSON.string_index_isvalid — Function.

Is i the index of the start of a character sequence in a JSON string s? A valid i can point to:

a single-byte (7-bit) character;

a '\X' escaped 7-bit character;

the start of a multi-byte UTF8 character;

the start of a \uXXXX escaped UTF16 character;

LazyJSON.tmp

LazyJSON.unescape_c — Function.

https://tools.ietf.org/html/rfc7159#section-7

char = unescaped /
    escape (
        %x22 /          ; "    quotation mark  U+0022
        %x5C /          ; \    reverse solidus U+005C
        %x2F /          ; /    solidus         U+002F
        %x62 /          ; b    backspace       U+0008
        %x66 /          ; f    form feed       U+000C
        %x6E /          ; n    line feed       U+000A
        %x72 /          ; r    carriage return U+000D
        %x74 /          ; t    tab             U+0009
        %x75 4HEXDIG )  ; uXXXX                U+XXXX

LazyJSON.unescape_hex4 — Function.

Read 4HEXDIG from a String s at byte index i + 1. FIXME more instructions but less memory access might be faster. Need to test.

LazyJSON.unescape_json_string — Function.

Unescape bytes of a String s up to byte index l, starting at byte index i. Return a new String.

LazyJSON.unhex_1

LazyJSON.unhex_10

LazyJSON.unhex_100

LazyJSON.unhex_1000

LazyJSON.unmangled_fieldnames

LazyJSON.value — Function.

value(jsontext) -> JSON.Value
JSON.Value <: Union{Number, AbstractString, AbstractVector, AbstractDict}

Create a JSON.Value object from a JSON text.

LazyJSON.wrap_object