Docstrings

Kuber.@K_str

Kuber.DEFAULT_NAMESPACE

Kuber.DEFAULT_URI

Kuber.KApi

Kuber.Kuber

Kuber.KuberContext

Kuber.Kubernetes

Kuber.Typedefs

Kuber._delopts

Kuber.api_group

Kuber.api_group_type

Kuber.api_method

Kuber.api_typedefs

Kuber.build_model_api_map

Kuber.camel

Kuber.connectDeleteNamespacedPodProxy

Kuber.connectDeleteNamespacedPodProxyWithPath

Kuber.connectDeleteNamespacedServiceProxy

Kuber.connectDeleteNamespacedServiceProxyWithPath

Kuber.connectDeleteNodeProxy

Kuber.connectDeleteNodeProxyWithPath

Kuber.connectGetNamespacedPodAttach

Kuber.connectGetNamespacedPodExec

Kuber.connectGetNamespacedPodPortforward

Kuber.connectGetNamespacedPodProxy

Kuber.connectGetNamespacedPodProxyWithPath

Kuber.connectGetNamespacedServiceProxy

Kuber.connectGetNamespacedServiceProxyWithPath

Kuber.connectGetNodeProxy

Kuber.connectGetNodeProxyWithPath

Kuber.connectHeadNamespacedPodProxy

Kuber.connectHeadNamespacedPodProxyWithPath

Kuber.connectHeadNamespacedServiceProxy

Kuber.connectHeadNamespacedServiceProxyWithPath

Kuber.connectHeadNodeProxy

Kuber.connectHeadNodeProxyWithPath

Kuber.connectOptionsNamespacedPodProxy

Kuber.connectOptionsNamespacedPodProxyWithPath

Kuber.connectOptionsNamespacedServiceProxy

Kuber.connectOptionsNamespacedServiceProxyWithPath

Kuber.connectOptionsNodeProxy

Kuber.connectOptionsNodeProxyWithPath

Kuber.connectPatchNamespacedPodProxy

Kuber.connectPatchNamespacedPodProxyWithPath

Kuber.connectPatchNamespacedServiceProxy

Kuber.connectPatchNamespacedServiceProxyWithPath

Kuber.connectPatchNodeProxy

Kuber.connectPatchNodeProxyWithPath

Kuber.connectPostNamespacedPodAttach

Kuber.connectPostNamespacedPodExec

Kuber.connectPostNamespacedPodPortforward

Kuber.connectPostNamespacedPodProxy

Kuber.connectPostNamespacedPodProxyWithPath

Kuber.connectPostNamespacedServiceProxy

Kuber.connectPostNamespacedServiceProxyWithPath

Kuber.connectPostNodeProxy

Kuber.connectPostNodeProxyWithPath

Kuber.connectPutNamespacedPodProxy

Kuber.connectPutNamespacedPodProxyWithPath

Kuber.connectPutNamespacedServiceProxy

Kuber.connectPutNamespacedServiceProxyWithPath

Kuber.connectPutNodeProxy

Kuber.connectPutNodeProxyWithPath

Kuber.createAPIService

Kuber.createCertificateSigningRequest

Kuber.createClusterRole

Kuber.createClusterRoleBinding

Kuber.createCustomResourceDefinition

Kuber.createInitializerConfiguration

Kuber.createMutatingWebhookConfiguration

Kuber.createNamespace

Kuber.createNamespacedBinding

Kuber.createNamespacedConfigMap

Kuber.createNamespacedControllerRevision

Kuber.createNamespacedCronJob

Kuber.createNamespacedDaemonSet

Kuber.createNamespacedDeployment

Kuber.createNamespacedDeploymentRollback

Kuber.createNamespacedEndpoints

Kuber.createNamespacedEvent

Kuber.createNamespacedHorizontalPodAutoscaler

Kuber.createNamespacedIngress

Kuber.createNamespacedJob

Kuber.createNamespacedLease

Kuber.createNamespacedLimitRange

Kuber.createNamespacedLocalSubjectAccessReview

Kuber.createNamespacedNetworkPolicy

Kuber.createNamespacedPersistentVolumeClaim

Kuber.createNamespacedPod

Kuber.createNamespacedPodBinding

Kuber.createNamespacedPodDisruptionBudget

Kuber.createNamespacedPodEviction

Kuber.createNamespacedPodPreset

Kuber.createNamespacedPodTemplate

Kuber.createNamespacedReplicaSet

Kuber.createNamespacedReplicationController

Kuber.createNamespacedResourceQuota

Kuber.createNamespacedRole

Kuber.createNamespacedRoleBinding

Kuber.createNamespacedSecret

Kuber.createNamespacedService

Kuber.createNamespacedServiceAccount

Kuber.createNamespacedStatefulSet

Kuber.createNode

Kuber.createPersistentVolume

Kuber.createPodSecurityPolicy

Kuber.createPriorityClass

Kuber.createSelfSubjectAccessReview

Kuber.createSelfSubjectRulesReview

Kuber.createStorageClass

Kuber.createSubjectAccessReview

Kuber.createTokenReview

Kuber.createValidatingWebhookConfiguration

Kuber.createVolumeAttachment

Kuber.delete!

Kuber.deleteAPIService

Kuber.deleteCertificateSigningRequest

Kuber.deleteClusterRole

Kuber.deleteClusterRoleBinding

Kuber.deleteCollectionAPIService

Kuber.deleteCollectionCertificateSigningRequest

Kuber.deleteCollectionClusterRole

Kuber.deleteCollectionClusterRoleBinding

Kuber.deleteCollectionCustomResourceDefinition

Kuber.deleteCollectionInitializerConfiguration

Kuber.deleteCollectionMutatingWebhookConfiguration

Kuber.deleteCollectionNamespacedConfigMap

Kuber.deleteCollectionNamespacedControllerRevision

Kuber.deleteCollectionNamespacedCronJob

Kuber.deleteCollectionNamespacedDaemonSet

Kuber.deleteCollectionNamespacedDeployment

Kuber.deleteCollectionNamespacedEndpoints

Kuber.deleteCollectionNamespacedEvent

Kuber.deleteCollectionNamespacedHorizontalPodAutoscaler

Kuber.deleteCollectionNamespacedIngress

Kuber.deleteCollectionNamespacedJob

Kuber.deleteCollectionNamespacedLease

Kuber.deleteCollectionNamespacedLimitRange

Kuber.deleteCollectionNamespacedNetworkPolicy

Kuber.deleteCollectionNamespacedPersistentVolumeClaim

Kuber.deleteCollectionNamespacedPod

Kuber.deleteCollectionNamespacedPodDisruptionBudget

Kuber.deleteCollectionNamespacedPodPreset

Kuber.deleteCollectionNamespacedPodTemplate

Kuber.deleteCollectionNamespacedReplicaSet

Kuber.deleteCollectionNamespacedReplicationController

Kuber.deleteCollectionNamespacedResourceQuota

Kuber.deleteCollectionNamespacedRole

Kuber.deleteCollectionNamespacedRoleBinding

Kuber.deleteCollectionNamespacedSecret

Kuber.deleteCollectionNamespacedServiceAccount

Kuber.deleteCollectionNamespacedStatefulSet

Kuber.deleteCollectionNode

Kuber.deleteCollectionPersistentVolume

Kuber.deleteCollectionPodSecurityPolicy

Kuber.deleteCollectionPriorityClass

Kuber.deleteCollectionStorageClass

Kuber.deleteCollectionValidatingWebhookConfiguration

Kuber.deleteCollectionVolumeAttachment

Kuber.deleteCustomResourceDefinition

Kuber.deleteInitializerConfiguration

Kuber.deleteMutatingWebhookConfiguration

Kuber.deleteNamespace

Kuber.deleteNamespacedConfigMap

Kuber.deleteNamespacedControllerRevision

Kuber.deleteNamespacedCronJob

Kuber.deleteNamespacedDaemonSet

Kuber.deleteNamespacedDeployment

Kuber.deleteNamespacedEndpoints

Kuber.deleteNamespacedEvent

Kuber.deleteNamespacedHorizontalPodAutoscaler

Kuber.deleteNamespacedIngress

Kuber.deleteNamespacedJob

Kuber.deleteNamespacedLease

Kuber.deleteNamespacedLimitRange

Kuber.deleteNamespacedNetworkPolicy

Kuber.deleteNamespacedPersistentVolumeClaim

Kuber.deleteNamespacedPod

Kuber.deleteNamespacedPodDisruptionBudget

Kuber.deleteNamespacedPodPreset

Kuber.deleteNamespacedPodTemplate

Kuber.deleteNamespacedReplicaSet

Kuber.deleteNamespacedReplicationController

Kuber.deleteNamespacedResourceQuota

Kuber.deleteNamespacedRole

Kuber.deleteNamespacedRoleBinding

Kuber.deleteNamespacedSecret

Kuber.deleteNamespacedService

Kuber.deleteNamespacedServiceAccount

Kuber.deleteNamespacedStatefulSet

Kuber.deleteNode

Kuber.deletePersistentVolume

Kuber.deletePodSecurityPolicy

Kuber.deletePriorityClass

Kuber.deleteStorageClass

Kuber.deleteValidatingWebhookConfiguration

Kuber.deleteVolumeAttachment

Kuber.eval

Kuber.fetch_api_machinery

Kuber.fetch_core_version

Kuber.fetch_misc_apis_versions

Kuber.get

Kuber.getAPIGroup

Kuber.getAPIResources

Kuber.Kubernetes.getAPIVersions — Function.

get available API versions Return: IoK8sApimachineryPkgApisMetaV1APIGroupList

Kuber.getCode

Kuber.get_logs — Function.

Fetch logs of the specified Pod

Parameters:

name::String (required)

Keyword Args:

container::String : The container for which to stream logs. Defaults to only container if there is one container in the pod.

follow::Bool : Follow the log stream of the pod. Defaults to false.

limitBytes::Int32 : If set, the number of bytes to read from the server before terminating the log output. This may not display a complete final line of logging, and may return slightly more or slightly less than the specified limit.

pretty::String : If 'true', then the output is pretty printed.

previous::Bool : Return previous terminated container logs. Defaults to false.

sinceSeconds::Int32 : A relative time in seconds before the current time from which to show logs. If this value precedes the time a pod was started, only logs since the pod start will be returned. If this value is in the future, no logs will be returned. Only one of sinceSeconds or sinceTime may be specified.

sinceTime::String : An RFC3339 timestamp from which to show logs. If this value precedes the time a pod was started, only logs since the pod start will be returned. If this value is in the future, no logs will be returned. Only one of sinceSeconds or sinceTime may be specified.

tailLines::Int32 : If set, the number of lines from the end of the logs to show. If not specified, logs are shown from the creation of the container or sinceSeconds or sinceTime

timestamps::Bool : If true, add an RFC3339 or RFC3339Nano timestamp at the beginning of every line of log output. Defaults to false.

Returns: String of all log entries, one per line

Kuber.get_ns

Kuber.get_server

Kuber.include

Kuber.is_json_mime

Kuber.kind_to_type

Kuber.kuber_obj

Kuber.kuber_type

Kuber.listAPIService

Kuber.listCertificateSigningRequest

Kuber.listClusterRole

Kuber.listClusterRoleBinding

Kuber.listComponentStatus

Kuber.listConfigMapForAllNamespaces

Kuber.listControllerRevisionForAllNamespaces

Kuber.listCronJobForAllNamespaces

Kuber.listCustomResourceDefinition

Kuber.listDaemonSetForAllNamespaces

Kuber.listDeploymentForAllNamespaces

Kuber.listEndpointsForAllNamespaces

Kuber.listEventForAllNamespaces

Kuber.listHorizontalPodAutoscalerForAllNamespaces

Kuber.listIngressForAllNamespaces

Kuber.listInitializerConfiguration

Kuber.listJobForAllNamespaces

Kuber.listLeaseForAllNamespaces

Kuber.listLimitRangeForAllNamespaces

Kuber.listMutatingWebhookConfiguration

Kuber.listNamespace

Kuber.listNamespacedConfigMap

Kuber.listNamespacedControllerRevision

Kuber.listNamespacedCronJob

Kuber.listNamespacedDaemonSet

Kuber.listNamespacedDeployment

Kuber.listNamespacedEndpoints

Kuber.listNamespacedEvent

Kuber.listNamespacedHorizontalPodAutoscaler

Kuber.listNamespacedIngress

Kuber.listNamespacedJob

Kuber.listNamespacedLease

Kuber.listNamespacedLimitRange

Kuber.listNamespacedNetworkPolicy

Kuber.listNamespacedPersistentVolumeClaim

Kuber.listNamespacedPod

Kuber.listNamespacedPodDisruptionBudget

Kuber.listNamespacedPodPreset

Kuber.listNamespacedPodTemplate

Kuber.listNamespacedReplicaSet

Kuber.listNamespacedReplicationController

Kuber.listNamespacedResourceQuota

Kuber.listNamespacedRole

Kuber.listNamespacedRoleBinding

Kuber.listNamespacedSecret

Kuber.listNamespacedService

Kuber.listNamespacedServiceAccount

Kuber.listNamespacedStatefulSet

Kuber.listNetworkPolicyForAllNamespaces

Kuber.listNode

Kuber.listPersistentVolume

Kuber.listPersistentVolumeClaimForAllNamespaces

Kuber.listPodDisruptionBudgetForAllNamespaces

Kuber.listPodForAllNamespaces

Kuber.listPodPresetForAllNamespaces

Kuber.listPodSecurityPolicy

Kuber.listPodTemplateForAllNamespaces

Kuber.listPriorityClass

Kuber.listReplicaSetForAllNamespaces

Kuber.listReplicationControllerForAllNamespaces

Kuber.listResourceQuotaForAllNamespaces

Kuber.listRoleBindingForAllNamespaces

Kuber.listRoleForAllNamespaces

Kuber.listSecretForAllNamespaces

Kuber.listServiceAccountForAllNamespaces

Kuber.listServiceForAllNamespaces

Kuber.listStatefulSetForAllNamespaces

Kuber.listStorageClass

Kuber.listValidatingWebhookConfiguration

Kuber.listVolumeAttachment

Kuber.override_pref

Kuber.patchAPIService

Kuber.patchAPIServiceStatus

Kuber.patchCertificateSigningRequest

Kuber.patchCertificateSigningRequestStatus

Kuber.patchClusterRole

Kuber.patchClusterRoleBinding

Kuber.patchCustomResourceDefinition

Kuber.patchCustomResourceDefinitionStatus

Kuber.patchInitializerConfiguration

Kuber.patchMutatingWebhookConfiguration

Kuber.patchNamespace

Kuber.patchNamespaceStatus

Kuber.patchNamespacedConfigMap

Kuber.patchNamespacedControllerRevision

Kuber.patchNamespacedCronJob

Kuber.patchNamespacedCronJobStatus

Kuber.patchNamespacedDaemonSet

Kuber.patchNamespacedDaemonSetStatus

Kuber.patchNamespacedDeployment

Kuber.patchNamespacedDeploymentScale

Kuber.patchNamespacedDeploymentStatus

Kuber.patchNamespacedEndpoints

Kuber.patchNamespacedEvent

Kuber.patchNamespacedHorizontalPodAutoscaler

Kuber.patchNamespacedHorizontalPodAutoscalerStatus

Kuber.patchNamespacedIngress

Kuber.patchNamespacedIngressStatus

Kuber.patchNamespacedJob

Kuber.patchNamespacedJobStatus

Kuber.patchNamespacedLease

Kuber.patchNamespacedLimitRange

Kuber.patchNamespacedNetworkPolicy

Kuber.patchNamespacedPersistentVolumeClaim

Kuber.patchNamespacedPersistentVolumeClaimStatus

Kuber.patchNamespacedPod

Kuber.patchNamespacedPodDisruptionBudget

Kuber.patchNamespacedPodDisruptionBudgetStatus

Kuber.patchNamespacedPodPreset

Kuber.patchNamespacedPodStatus

Kuber.patchNamespacedPodTemplate

Kuber.patchNamespacedReplicaSet

Kuber.patchNamespacedReplicaSetScale

Kuber.patchNamespacedReplicaSetStatus

Kuber.patchNamespacedReplicationController

Kuber.patchNamespacedReplicationControllerDummyScale

Kuber.patchNamespacedReplicationControllerScale

Kuber.patchNamespacedReplicationControllerStatus

Kuber.patchNamespacedResourceQuota

Kuber.patchNamespacedResourceQuotaStatus

Kuber.patchNamespacedRole

Kuber.patchNamespacedRoleBinding

Kuber.patchNamespacedSecret

Kuber.patchNamespacedService

Kuber.patchNamespacedServiceAccount

Kuber.patchNamespacedServiceStatus

Kuber.patchNamespacedStatefulSet

Kuber.patchNamespacedStatefulSetScale

Kuber.patchNamespacedStatefulSetStatus

Kuber.patchNode

Kuber.patchNodeStatus

Kuber.patchPersistentVolume

Kuber.patchPersistentVolumeStatus

Kuber.patchPodSecurityPolicy

Kuber.patchPriorityClass

Kuber.patchStorageClass

Kuber.patchValidatingWebhookConfiguration

Kuber.patchVolumeAttachment

Kuber.put!

Kuber.readAPIService

Kuber.readAPIServiceStatus

Kuber.readCertificateSigningRequest

Kuber.readCertificateSigningRequestStatus

Kuber.readClusterRole

Kuber.readClusterRoleBinding

Kuber.readComponentStatus

Kuber.readCustomResourceDefinition

Kuber.readCustomResourceDefinitionStatus

Kuber.readInitializerConfiguration

Kuber.readMutatingWebhookConfiguration

Kuber.readNamespace

Kuber.readNamespaceStatus

Kuber.readNamespacedConfigMap

Kuber.readNamespacedControllerRevision

Kuber.readNamespacedCronJob

Kuber.readNamespacedCronJobStatus

Kuber.readNamespacedDaemonSet

Kuber.readNamespacedDaemonSetStatus

Kuber.readNamespacedDeployment

Kuber.readNamespacedDeploymentScale

Kuber.readNamespacedDeploymentStatus

Kuber.readNamespacedEndpoints

Kuber.readNamespacedEvent

Kuber.readNamespacedHorizontalPodAutoscaler

Kuber.readNamespacedHorizontalPodAutoscalerStatus

Kuber.readNamespacedIngress

Kuber.readNamespacedIngressStatus

Kuber.readNamespacedJob

Kuber.readNamespacedJobStatus

Kuber.readNamespacedLease

Kuber.readNamespacedLimitRange

Kuber.readNamespacedNetworkPolicy

Kuber.readNamespacedPersistentVolumeClaim

Kuber.readNamespacedPersistentVolumeClaimStatus

Kuber.readNamespacedPod

Kuber.readNamespacedPodDisruptionBudget

Kuber.readNamespacedPodDisruptionBudgetStatus

Kuber.readNamespacedPodLog

Kuber.readNamespacedPodPreset

Kuber.readNamespacedPodStatus

Kuber.readNamespacedPodTemplate

Kuber.readNamespacedReplicaSet

Kuber.readNamespacedReplicaSetScale

Kuber.readNamespacedReplicaSetStatus

Kuber.readNamespacedReplicationController

Kuber.readNamespacedReplicationControllerDummyScale

Kuber.readNamespacedReplicationControllerScale

Kuber.readNamespacedReplicationControllerStatus

Kuber.readNamespacedResourceQuota

Kuber.readNamespacedResourceQuotaStatus

Kuber.readNamespacedRole

Kuber.readNamespacedRoleBinding

Kuber.readNamespacedSecret

Kuber.readNamespacedService

Kuber.readNamespacedServiceAccount

Kuber.readNamespacedServiceStatus

Kuber.readNamespacedStatefulSet

Kuber.readNamespacedStatefulSetScale

Kuber.readNamespacedStatefulSetStatus

Kuber.readNode

Kuber.readNodeStatus

Kuber.readPersistentVolume

Kuber.readPersistentVolumeStatus

Kuber.readPodSecurityPolicy

Kuber.readPriorityClass

Kuber.readStorageClass

Kuber.readValidatingWebhookConfiguration

Kuber.readVolumeAttachment

Kuber.replaceAPIService

Kuber.replaceAPIServiceStatus

Kuber.replaceCertificateSigningRequest

Kuber.replaceCertificateSigningRequestApproval

Kuber.replaceCertificateSigningRequestStatus

Kuber.replaceClusterRole

Kuber.replaceClusterRoleBinding

Kuber.replaceCustomResourceDefinition

Kuber.replaceCustomResourceDefinitionStatus

Kuber.replaceInitializerConfiguration

Kuber.replaceMutatingWebhookConfiguration

Kuber.replaceNamespace

Kuber.replaceNamespaceFinalize

Kuber.replaceNamespaceStatus

Kuber.replaceNamespacedConfigMap

Kuber.replaceNamespacedControllerRevision

Kuber.replaceNamespacedCronJob

Kuber.replaceNamespacedCronJobStatus

Kuber.replaceNamespacedDaemonSet

Kuber.replaceNamespacedDaemonSetStatus

Kuber.replaceNamespacedDeployment

Kuber.replaceNamespacedDeploymentScale

Kuber.replaceNamespacedDeploymentStatus

Kuber.replaceNamespacedEndpoints

Kuber.replaceNamespacedEvent

Kuber.replaceNamespacedHorizontalPodAutoscaler

Kuber.replaceNamespacedHorizontalPodAutoscalerStatus

Kuber.replaceNamespacedIngress

Kuber.replaceNamespacedIngressStatus

Kuber.replaceNamespacedJob

Kuber.replaceNamespacedJobStatus

Kuber.replaceNamespacedLease

Kuber.replaceNamespacedLimitRange

Kuber.replaceNamespacedNetworkPolicy

Kuber.replaceNamespacedPersistentVolumeClaim

Kuber.replaceNamespacedPersistentVolumeClaimStatus

Kuber.replaceNamespacedPod

Kuber.replaceNamespacedPodDisruptionBudget

Kuber.replaceNamespacedPodDisruptionBudgetStatus

Kuber.replaceNamespacedPodPreset

Kuber.replaceNamespacedPodStatus

Kuber.replaceNamespacedPodTemplate

Kuber.replaceNamespacedReplicaSet

Kuber.replaceNamespacedReplicaSetScale

Kuber.replaceNamespacedReplicaSetStatus

Kuber.replaceNamespacedReplicationController

Kuber.replaceNamespacedReplicationControllerDummyScale

Kuber.replaceNamespacedReplicationControllerScale

Kuber.replaceNamespacedReplicationControllerStatus

Kuber.replaceNamespacedResourceQuota

Kuber.replaceNamespacedResourceQuotaStatus

Kuber.replaceNamespacedRole

Kuber.replaceNamespacedRoleBinding

Kuber.replaceNamespacedSecret

Kuber.replaceNamespacedService

Kuber.replaceNamespacedServiceAccount

Kuber.replaceNamespacedServiceStatus

Kuber.replaceNamespacedStatefulSet

Kuber.replaceNamespacedStatefulSetScale

Kuber.replaceNamespacedStatefulSetStatus

Kuber.replaceNode

Kuber.replaceNodeStatus

Kuber.replacePersistentVolume

Kuber.replacePersistentVolumeStatus

Kuber.replacePodSecurityPolicy

Kuber.replacePriorityClass

Kuber.replaceStorageClass

Kuber.replaceValidatingWebhookConfiguration

Kuber.replaceVolumeAttachment

Kuber.sel

Kuber.set_api_versions!

Kuber.set_ns

Kuber.set_server

Kuber.update!

Kuber.watchAPIService

Kuber.watchAPIServiceList

Kuber.watchCertificateSigningRequest

Kuber.watchCertificateSigningRequestList

Kuber.watchClusterRole

Kuber.watchClusterRoleBinding

Kuber.watchClusterRoleBindingList

Kuber.watchClusterRoleList

Kuber.watchConfigMapListForAllNamespaces

Kuber.watchControllerRevisionListForAllNamespaces

Kuber.watchCronJobListForAllNamespaces

Kuber.watchCustomResourceDefinition

Kuber.watchCustomResourceDefinitionList

Kuber.watchDaemonSetListForAllNamespaces

Kuber.watchDeploymentListForAllNamespaces

Kuber.watchEndpointsListForAllNamespaces

Kuber.watchEventListForAllNamespaces

Kuber.watchHorizontalPodAutoscalerListForAllNamespaces

Kuber.watchIngressListForAllNamespaces

Kuber.watchInitializerConfiguration

Kuber.watchInitializerConfigurationList

Kuber.watchJobListForAllNamespaces

Kuber.watchLeaseListForAllNamespaces

Kuber.watchLimitRangeListForAllNamespaces

Kuber.watchMutatingWebhookConfiguration

Kuber.watchMutatingWebhookConfigurationList

Kuber.watchNamespace

Kuber.watchNamespaceList

Kuber.watchNamespacedConfigMap

Kuber.watchNamespacedConfigMapList

Kuber.watchNamespacedControllerRevision

Kuber.watchNamespacedControllerRevisionList

Kuber.watchNamespacedCronJob

Kuber.watchNamespacedCronJobList

Kuber.watchNamespacedDaemonSet

Kuber.watchNamespacedDaemonSetList

Kuber.watchNamespacedDeployment

Kuber.watchNamespacedDeploymentList

Kuber.watchNamespacedEndpoints

Kuber.watchNamespacedEndpointsList

Kuber.watchNamespacedEvent

Kuber.watchNamespacedEventList

Kuber.watchNamespacedHorizontalPodAutoscaler

Kuber.watchNamespacedHorizontalPodAutoscalerList

Kuber.watchNamespacedIngress

Kuber.watchNamespacedIngressList

Kuber.watchNamespacedJob

Kuber.watchNamespacedJobList

Kuber.watchNamespacedLease

Kuber.watchNamespacedLeaseList

Kuber.watchNamespacedLimitRange

Kuber.watchNamespacedLimitRangeList

Kuber.watchNamespacedNetworkPolicy

Kuber.watchNamespacedNetworkPolicyList

Kuber.watchNamespacedPersistentVolumeClaim

Kuber.watchNamespacedPersistentVolumeClaimList

Kuber.watchNamespacedPod

Kuber.watchNamespacedPodDisruptionBudget

Kuber.watchNamespacedPodDisruptionBudgetList

Kuber.watchNamespacedPodList

Kuber.watchNamespacedPodPreset

Kuber.watchNamespacedPodPresetList

Kuber.watchNamespacedPodTemplate

Kuber.watchNamespacedPodTemplateList

Kuber.watchNamespacedReplicaSet

Kuber.watchNamespacedReplicaSetList

Kuber.watchNamespacedReplicationController

Kuber.watchNamespacedReplicationControllerList

Kuber.watchNamespacedResourceQuota

Kuber.watchNamespacedResourceQuotaList

Kuber.watchNamespacedRole

Kuber.watchNamespacedRoleBinding

Kuber.watchNamespacedRoleBindingList

Kuber.watchNamespacedRoleList

Kuber.watchNamespacedSecret

Kuber.watchNamespacedSecretList

Kuber.watchNamespacedService

Kuber.watchNamespacedServiceAccount

Kuber.watchNamespacedServiceAccountList

Kuber.watchNamespacedServiceList

Kuber.watchNamespacedStatefulSet

Kuber.watchNamespacedStatefulSetList

Kuber.watchNetworkPolicyListForAllNamespaces

Kuber.watchNode

Kuber.watchNodeList

Kuber.watchPersistentVolume

Kuber.watchPersistentVolumeClaimListForAllNamespaces

Kuber.watchPersistentVolumeList

Kuber.watchPodDisruptionBudgetListForAllNamespaces

Kuber.watchPodListForAllNamespaces

Kuber.watchPodPresetListForAllNamespaces

Kuber.watchPodSecurityPolicy

Kuber.watchPodSecurityPolicyList

Kuber.watchPodTemplateListForAllNamespaces

Kuber.watchPriorityClass

Kuber.watchPriorityClassList

Kuber.watchReplicaSetListForAllNamespaces

Kuber.watchReplicationControllerListForAllNamespaces

Kuber.watchResourceQuotaListForAllNamespaces

Kuber.watchRoleBindingListForAllNamespaces

Kuber.watchRoleListForAllNamespaces

Kuber.watchSecretListForAllNamespaces

Kuber.watchServiceAccountListForAllNamespaces

Kuber.watchServiceListForAllNamespaces

Kuber.watchStatefulSetListForAllNamespaces

Kuber.watchStorageClass

Kuber.watchStorageClassList

Kuber.watchValidatingWebhookConfiguration

Kuber.watchValidatingWebhookConfigurationList

Kuber.watchVolumeAttachment

Kuber.watchVolumeAttachmentList