Docstrings

YAJL.@yajl — Macro.

Register a callback for a specific data type. Callback functions should return true or a non-zero integer upon success, otherwise parsing stops (although sometimes this is desired). A return true is inserted automatically at the end of the function, but your own explicit returns override this. Note that the return keyword must be used in this case.

The callbacks to be overridden are as follows:

null(ctx::T): Called on null values.

boolean(ctx::T, v::Bool): `Called on boolean values.

integer(ctx::T, v::Int): Called on integer values (see note below).

double(ctx::T, v::Float64): Called on float values (see note below).

number(ctx::T, v::Ptr{UInt8}, len::Int): Called on numeric values (see note below).

string(ctx::T, v::Ptr{UInt8}, len::Int): Called on string values.

map_start(ctx::T): Called when an object begins ({).

map_key(ctx::T, v::Ptr{UInt8}, len::Int): Called on object keys.

map_end(ctx::T): Called when an object ends (}).

array_start(ctx::T): Called when an array begins ([).

array_end(ctx::T): Called when an array ends (]).

For string arguments which appear as Ptr{UInt8}, Cstring can also be used. However, Ptr{UInt8} is usually better if you want to use unsafe_string(v, len).

Note

To handle numbers, implement either number or both integer and double. Usually, number is a better choice because integer and `double have limited precision. See here for more details.

Warning

If your Context is a parametric type, it must appear non-parameterized in the function definition. This means that your callback functions cannot dispatch on the context's type parameter.

For a full example, see minifier.jl.

YAJL.ALLOW_COMMENTS

YAJL.ALLOW_MULTIPLE_VALUES

YAJL.ALLOW_PARTIAL_VALUES

YAJL.ALLOW_TRAILING_GARBAGE

YAJL.CALLBACKS

YAJL.CLOSE_BRACE

YAJL.CLOSE_BRACKET

YAJL.COLON

YAJL.COMMA

YAJL.Callbacks

YAJL.Context — Type.

Base type for YAJL contexts. To implement a custom Context's behaviour, see @yajl and collect. For a full example, see minifier.jl.

YAJL.DONT_VALIDATE_STRINGS

YAJL.MINIFY_ARRAY

YAJL.MINIFY_ARRAY_FIRST

YAJL.MINIFY_INIT

YAJL.MINIFY_MAP_KEY

YAJL.MINIFY_MAP_KEY_FIRST

YAJL.MINIFY_MAP_VAL

YAJL.Minifier — Type.

Minifier(io::IO=stdout) -> Minifier

Removes all unecessary whitespace from JSON.

Example

julia> String(take!(YAJL.run(IOBuffer("{    }"), YAJL.Minifier(IOBuffer()))))
"{}"

YAJL.MinifyState

YAJL.NULL

YAJL.OPEN_BRACE

YAJL.OPEN_BRACKET

YAJL.OPTIONS

YAJL.ParseError

YAJL.QUOTE

YAJL.ST_CLIENT_CANCELLED

YAJL.ST_ERROR

YAJL.ST_OK

YAJL.YAJL

YAJL.__init__

YAJL.addval

YAJL.cb_array_end

YAJL.cb_array_start

YAJL.cb_boolean

YAJL.cb_double

YAJL.cb_integer

YAJL.cb_map_end

YAJL.cb_map_key

YAJL.cb_map_start

YAJL.cb_null

YAJL.cb_number

YAJL.cb_string

YAJL.check_deps

YAJL.checkstatus

YAJL.checktypes

YAJL.collect — Function.

collect(ctx::Context)

Override this function for your custom Context to specify what is returned from run. By default, ctx itself is returned.

YAJL.depsfile

YAJL.eval

YAJL.hasmethod

YAJL.include

YAJL.libyajl

YAJL.on_array_end

YAJL.on_array_start

YAJL.on_boolean

YAJL.on_map_end

YAJL.on_map_key

YAJL.on_map_start

YAJL.on_null

YAJL.on_number

YAJL.on_string

YAJL.run — Function.

run(io::IO, ctx::Context; chunk::Integer=2^16, options::Integer=0x0)

Parse the JSON data from io and process it with ctx's callbacks. The return value is determined by the implementation of collect for ctx.

Keywords

chunk::Integer=2^16: Number of bytes to read from io at a time.

options::Integer=0x0: YAJL parser options, ORed together.

YAJL.writeval

YAJL.yajl