Docstrings

ScHoLP.HONData — Type.

HONData

Data structure for a temporal higher-order network.

Each dataset consists of three integer vectors: simplices, nverts, and times.

The simplices is a contiguous vector of nodes comprising the simplices.

The nverts vector contains the number of vertices within each simplex.

The times vector contains the timestamps of the simplices (same length as nverts).

For example, consider a dataset consisting of three simplices:

1. {1, 2, 3} at time 10
2. {2, 4} at time 15.
3. {1, 3, 4, 5} at time 21.

Then the data structure would be

simplices = [1, 2, 3, 2, 4, 1, 3, 4, 5]

nverts = [3, 2, 4]

times = [10, 15, 21]

There is an additional name variable attached to the dataset.

ScHoLP.NbrSetMap — Type.

NbrSetMap

const NbrSetMap = Dict{NTuple{2, Int64}, Set{Int64}}

ScHoLP.OPEN

ScHoLP.PKatz3 — Function.

PKatz3

Compute 3-way personalized Katz scores for triangles.

PKatz3(triangles::Vector{NTuple{3,Int64}}, B::SpIntMat, unweighted::Bool, dense_solve::Bool=false)

Example usage:

(scores, S) = PPR3(T, B, false, false)

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

B::SpIntMat: Projected graph as a Sparse integer matrix, where B[i, j] is the number of times that i and j co-appear in a simplex.

unweighted::Bool: Whether or not to use the unweighted version of the matrix.

dense_solve::Bool=false: whether or not to use a dense solver. If B is small, then it is worth setting this option to true.

returns a tuple (scores, S):

scores::Vector{Float64}: a vector of 3-way PPR scores for the triangles

S::SpFltMat: a sparse matrix with the same sparsity pattern as B, where S[i, j] is the personalized Katz score of node i with respect to the seed node j.

ScHoLP.PPR3 — Function.

PPR3

Compute 3-way personalized PageRank scores for triangles.

PPR3(triangles::Vector{NTuple{3,Int64}}, B::SpIntMat, unweighted::Bool, dense_solve::Bool=false, α::Float64=0.85)

Example usage:

(scores, S) = PPR3(T, B, false, false, 0.85)

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

B::SpIntMat: Projected graph as a Sparse integer matrix, where B[i, j] is the number of times that i and j co-appear in a simplex.

unweighted::Bool: Whether or not to use the unweighted version of the matrix.

dense_solve::Bool=false: whether or not to use a dense solver. If B is small, then it is worth setting this option to true.

α::Float64=0.85: teleportation parameter for PageRank

returns a tuple (scores, S):

scores::Vector{Float64}: a vector of 3-way PPR scores for the triangles

S::SpFltMat: a sparse matrix with the same sparsity pattern as B, where S[i, j] is the personalized PageRank score of node i with respect to the seed node j.

ScHoLP.STRONG

ScHoLP.ScHoLP

ScHoLP.SimplicialPROperator — Function.

SimplicialPROperator

Construct the Simplicial PageRank Operator

SimplicialPROperator(grad::SpIntMat, curl::SpIntMat, edge_map::Dict{NTuple{2,Int64},Int64}, α::Float64)

Input parameters:

grad::SpIntMat: gradient operator (as a matrix)

curl::SpIntMat: curl operator (as a matrix)

edge_map::Dict{NTuple{2,Int64}, Int64}: maps an a sorted edge tuple to an index for the matrices

α::Float64: teleportation parameter

returns simplicial PageRank operator

ScHoLP.Simplicial_PPR3_combined — Function.

SimplicialPPR3combined

Compute the "combined" 3-way simplicial personalized PageRank scores for triangles. This is less computationally expensive that the related function SimplicialPPR3decomposed(), which decomposes the scores into the curl, gradient, and harmonic components.

function SimplicialPPR3combined(triangles::Vector{NTuple{3,Int64}}, A::SpIntMat, At::SpIntMat, B::SpIntMat, α::Float64=0.85)

Example usage:

(scores_comb, S_comb, vec_edge_map) = Simplicial_PPR3_combined(T, A)

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

A::SpIntMat: (# nodes) x (# simplices) adjacency matrix

α::Float64=0.85: teleportation parameter for PageRank

returns a tuple (scorescomb, Scomb, vecedgemap)

scores_comb::Vector{Float64}: scores for "combined" simplicial personalized PageRank

S_comb::SpFltMat: a sparse matrix, where S[x, y] is the combined simplicial PPR score for edge x with respect to edge y

vecedgemap::Array{Int64,2}: 2 x (# edges) map of indices for Scomb such that the vecedge_map[:, i] is the edge for ith index in the matrix

ScHoLP.Simplicial_PPR3_decomposed — Function.

SimplicialPPR3decomposed

Compute 3-way simplicial personalized PageRank scores for triangles, decomposed into the harmonic, gradient, and curl components.

SimplicialPPR3decomposed(triangles::Vector{NTuple{3,Int64}}, A::SpIntMat, dense_solve::Bool=false, α::Float64=0.85)

Example usage:

(scores_comb, scores_curl, scores_grad, scores_harm, S_comb, S_curl, S_grad, S_harm, vec_edge_map) =
    Simplicial_PPR3_decomposed(T, A)

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

A::SpIntMat: (# nodes) x (# simplices) adjacency matrix

dense_solve::Bool=false: whether or not to use a dense solver. If the network is tiny, then it might be worth setting this option to true.

α::Float64=0.85: teleportation parameter for PageRank

returns a tuple (scorescomb, scorescurl, scoresgrad, scoresharm, Scomb, Scurl, Sgrad, Sharm, vecedgemap)

scores_comb::Vector{Float64}: scores for "combined" simplicial personalized PageRank

scores_curl::Vector{Float64}: scores for curl component

scores_grad::Vector{Float64}: scores for gradient component

scores_harm::Vector{Float64}: scores for harmonic component

S_comb::SpFltMat: a sparse matrix, where S[x, y] is the combined simplicial PPR score for edge x with respect to edge y

Scurl::SpFltMat: same as Scomb but for the curl component

Sgrad::SpFltMat: same as Scomb but for the gradient component

Sharm::SpFltMat: same as Scomb but for the harmonic component

vecedgemap::Array{Int64,2}: 2 x (# edges) map of indices for Scomb, Scurl, Sgrad, Sharm, such that the vecedgemap[:, i] is the edge for ith index in the matrix

ScHoLP.SpFltMat — Type.

SpFltMat

const SpFltMat = SparseMatrixCSC{Float64,Int64}

ScHoLP.SpIntMat — Type.

SpIntMat

const SpIntMat = SparseMatrixCSC{Float64,Int64}

ScHoLP.SpMat — Constant.

SpMat

const SpMat = Union{SpIntMat,SpFltMat}

ScHoLP.WEAK

ScHoLP.adamic_adar3 — Function.

adamic_adar3

Returns the 3-way Adamic-Adar score for a list of triangles.

adamicadar3(triangles::Vector{NTuple{3,Int64}}, commonnbrs::NbrSetMap, degrees::Vector{Int64})

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

commonnbrs::NbrSetMap: the common neighbors map attained from commonneighbors_map()

degrees::Vector{Int64}: degree of each node in the projected graph

ScHoLP.arithmetic_mean — Function.

arithmetic_mean

Returns the arithmetic mean of the weights of the edges of a list of triangles.

arithmetic_mean(triangles::Vector{NTuple{3,Int64}}, B::SpIntMat)

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

B::SpIntMat: Projected graph as a Sparse integer matrix, where B[i, j] is the number of times that i and j co-appear in a simplex.

ScHoLP.backbone

ScHoLP.basic_matrices — Function.

basic_matrices

Computes some simple matrices associated with a dataset.

basic_matrices(simplices::Vector{Int64}, nverts::Vector{Int64})

Input parameters:

simplices::Vector{Int64}: the contiguous vector of simplices

nverts::Vector{Int64}: the vector of sizes of simplices

Outputs tuple (A, At, B):

A::SpIntMat: (# nodes) x (# simplices) adjacency matrix

At::SpIntMat: the transpose of A

B::SpIntMat: Projected graph as a Sparse integer matrix, where B[i, j] is the number of times that i and j co-appear in a simplex.

basic_matrices

Computes some simple matrices associated with a dataset.

basic_matrices(dataset::HONData)

Input parameter:

dataset::HONData: the dataset

Outputs tuple (A, At, B):

A::SpIntMat: (# nodes) x (# simplices) adjacency matrix

At::SpIntMat: the transpose of A

B::SpIntMat: Projected graph as a Sparse integer matrix, where B[i, j] is the number of times that i and j co-appear in a simplex.

ScHoLP.basic_summary_statistics — Function.

basicsummarystatistics

Prints some basic summary statistics of the dataset.

basicsummarystatistics(dataset::String)

Input parameter:

dataset::HONData: The dataset.

ScHoLP.bipartite_graph

ScHoLP.closure_type_counts3 — Function.

closuretypecounts3

Computes the closure probabilities of all 3-node configurations. The closure probability is the fraction of instances of open configurations appearing in the first 80% of the timestamped simplices that appear in a simplex in the final 20%.

closuretypecounts3(dataset::HONData, initial_cutoff::Int64=100)

Input parameters:

dataset::String: The dataset name.

initialcutoff::Int64=100: Initial cutoff of the simplices. If this is set to less than 100, then the data is first preprocessed to only consider the first initialcutoff percentage of the data.

ScHoLP.closure_type_counts4 — Function.

closuretypecounts4

Computes the closure probabilities of all 4-node configurations that contain at least one triangle. The closure probability is the fraction of instances of open configurations appearing in the first 80% of the timestamped simplices that appear in a simplex in the final 20%.

closuretypecounts4(dataset::HONData, initial_cutoff::Int64=100)

Input parameters:

dataset::HONData: The dataset.

initialcutoff::Int64=100: Initial cutoff of the simplices. If this is set to less than 100, then the data is first preprocessed to only consider the first initialcutoff percentage of the data.

ScHoLP.collect_lifecycles

ScHoLP.common3 — Function.

common3

Returns the number of common 4th neighbors of a list of triangles.

common3(triangles::Vector{NTuple{3,Int64}}, common_nbrs::NbrSetMap)

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

commonnbrs::NbrSetMap: the common neighbors map attained from commonneighbors_map()

ScHoLP.common_nbr_set — Function.

Return common neighbors of two nodes u and v.

ScHoLP.common_neighbors_map — Function.

commonneighborsmap

Construct a map where a key is an edge in the graph B participating in at least one triangle of interest and a value is the vector of common neighbors of the end points of the edge. The graph B is assumed to be undirected, and the keys are ordered by the pair with smallest ID first.

commonneighborsmap(B::SpIntMat, triangles::Vector{NTuple{3,Int64}})

Input parameters:

B::SpIntMat: the graph

triangles::Vector{NTuple{3,Int64}}: the triangles of interest

ScHoLP.configuration_sizes_preserved

ScHoLP.data_size_cutoff

ScHoLP.earliest_activity

ScHoLP.enum_open_triangles

ScHoLP.eval

ScHoLP.example_dataset — Function.

example_dataset

Returns one of the example datasets.

example_dataset(dataset::String)

Input parameter:

dataset::String: one of "email-Enron", "contact-primary-school", "contact-high-school", "example1", or "example2"

ScHoLP.full_solve

ScHoLP.generalized_means — Function.

generalized_means

Computes the generalized p-means of the weights of the edges of a list of triangles. The generalized mean of 3 values is

\[M_p(x, y, z) = ((x^p + y^p + z^p) / 3)^{1/p}\]

generalized_means(triangles::Vector{NTuple{3,Int64}}, B::SpIntMat, ps::Float64=[-Inf; collect(-4:0.25:4); Inf])

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

B::SpIntMat: Projected graph as a Sparse integer matrix, where B[i, j] is the number of times that i and j co-appear in a simplex.

ps::Vector{Float64}=[-Inf; collect(-4:0.25:4); Inf]: the values of p for which to compute the means

Returns a matrix of size length(triangles) x length(ps) of the scores for the various generalized means.

ScHoLP.geometric_mean — Function.

geometric_mean

Returns the geometric mean of the weights of the edges of a list of triangles.

geometric_mean(triangles::Vector{NTuple{3,Int64}}, B::SpIntMat)

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

B::SpIntMat: Projected graph as a Sparse integer matrix, where B[i, j] is the number of times that i and j co-appear in a simplex.

ScHoLP.grad_and_curl — Function.

gradandcurl

Construct the gradient and curl operators.

gradandcurl(A::SpIntMat, At::SpIntMat, B::SpIntMat)

Input parameters:

A::SpIntMat: (# nodes) x (# simplices) adjacency matrix

At::SpIntMat: the transpose of A

B::SpIntMat: Projected graph as a Sparse integer matrix, where B[i, j] is the number of times that i and j co-appear in a simplex.

returns tuple (grad, curl, edge_map):

grad::SpIntMat: gradient operator (as a matrix)

curl::SpIntMat: curl operator (as a matrix)

edge_map::Dict{NTuple{2,Int64}, Int64}: maps an a sorted edge tuple to an index for the matrices

ScHoLP.harmonic_mean — Function.

harmonic_mean

Returns the harmonic mean of the weights of the edges of a list of triangles.

harmonic_mean(triangles::Vector{NTuple{3,Int64}}, B::SpIntMat)

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

B::SpIntMat: Projected graph as a Sparse integer matrix, where B[i, j] is the number of times that i and j co-appear in a simplex.

ScHoLP.hodge_normalization

ScHoLP.include

ScHoLP.initialize_type_counter3

ScHoLP.initialize_type_counter4

ScHoLP.interval_overlaps — Function.

interval_overlaps

interval_overlaps(dataset::HONData)

Compute the number of active interval overlaps in open triangles.

dataset::HONData The dataset.

ScHoLP.intervals

ScHoLP.iterative_solve

ScHoLP.jaccard3 — Function.

jaccard3

Returns the 3-way Jaccard index of the neighbor lists of nodes in triangle.

jaccard3(triangles::Vector{NTuple{3,Int64}}, common_nbrs::NbrSetMap, degrees::Vector{Int64})

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

commonnbrs::NbrSetMap: the common neighbors map attained from commonneighbors_map()

degrees::Vector{Int64}: degree of each node in the projected graph

ScHoLP.make_sparse_ones — Function.

makesparseones

Returns a new sparse matrix with the same non-zero pattern as the input but where all non-zeros are set to 1.

makesparseones(A::SpIntMat)

Input parameter:

A::SpIntMat: a sparse matrix

ScHoLP.neighbor_pairs

ScHoLP.neighbors

ScHoLP.new_closures

ScHoLP.newly_closed_types3

ScHoLP.newly_closed_types4

ScHoLP.num_open_closed_triangles — Function.

numopenclosed_triangles

Computes the number of open and closed triangles in a dataset.

numopenclosed_triangles(data::HONData)

Input parameter:

data::HONData: the dataset

Outputs tuple (no, nc):

no: number of open triangles

nc: number of closed triangles

ScHoLP.nz_row_inds

ScHoLP.nz_row_vals

ScHoLP.open_types3

ScHoLP.open_types4

ScHoLP.pref_attach3 — Function.

pref_attach3

Returns the preferential attachment score for a degree vector

pref_attach3(triangles::Vector{NTuple{3,Int64}}, degrees::Vector{Int64})

The score of triangle (i, j, k) is degrees[i] * degrees[j] * degrees[k]

Input parameters:

triangles::Vector{NTuple{3,Int64}}: The vector of triangles upon which to compute scores.

degrees::Vector{Int64}: the degree vector

ScHoLP.process_lifecycles — Function.

process_lifecycles

Process all lifecycles in a given dataset.

process_lifecycles(data::HONData)

Input parameters:

data::HONData: The dataset

Returns a tuple (closedtransitions, opentransitions)

closed_transitions::Array{Int64,2}: matrix whose (i, j) entry is the number of transitions from configuration j to configuration i, out of all triples of nodes that eventually simplicially close

open_transitions::Array{Int64,2}: matrix whose (i, j) entry is the number of transitions from configuration j to configuration i, out of all triples of nodes that do not simplicially close

ScHoLP.process_trail

ScHoLP.proj_graph_degree_order — Function.

Ordering of nodes by their degree

ScHoLP.remove_diagonal

ScHoLP.simplex_degree_order — Function.

Ordering of nodes by the number of simplices in which they appear

ScHoLP.simplex_key3

ScHoLP.simplex_key4

ScHoLP.sorted_tuple — Function.

Turns 3 integers into a sorted tuple.

Turns 4 integers into a sorted tuple.

ScHoLP.split_data — Function.

split_data

Split data by timestamps into quantiles specified by quantile1 and quantile2. Returns a 4-tuple (oldsimps, oldnverts, newsimps, newnverts), where (oldsimps, oldnverts) are the data in the quantile [0, quantile1] and (newsimps, newnverts) are the data in the quantile (quantile1, quantile2].

split_data(simplices::Vector{Int64}, nverts::Vector{Int64}, times::Vector{Int64}, quantile1::Int64, quantile2::Int64)

Input parameters:

simplices::Vector{Int64}: contiguous vector of simplices from dataset

nverts::Vector{Int64}: vector of simplex sizes from dataset

times::Vector{Int64}: vector of timestamps of simplices

quantile1::Int64: first quantile

quantile2::Int64: second quantile

Returns a tuple (oldsimps, oldnverts, newsimps, newnverts):

old_simplices::Vector{Int64}: simplices in quantile1

oldnverts::Vector{Int64}: sizes of simplices in oldsimplices

new_simplices::Vector{Int64}: simplices between quantile1 and quantile 2

newnverts::Vector{Int64}: sizes of simplices in newsimplices

ScHoLP.summary_statistics — Function.

summary_statistics

Computes several statistics about the dataset.

summary_statistics(dataset::HONData)

Input parameter:

dataset::HONData: The dataset.

ScHoLP.tetrahedron_closed

ScHoLP.triangle_closed

ScHoLP.triangle_counts — Function.

triangle_counts

Counts the number of times each simplicial triangle appears in the data. This function is expensive and requires storage on the order of the number of triangles. Returns a counter that takes a sorted triple of indices as a key and returns the count of that triangle.

triangle_counts(data::HONData)

Input parameters:

data::HONData: The data.

ScHoLP.window_data