Docstrings

NetworkInference.AbstractNetworkInference

NetworkInference.CLRNetworkInference

NetworkInference.EB_EXISTS

NetworkInference.Edge — Type.

Undirected edge

Fields:

nodes: the two nodes, in an arbitrary order

weight: weight indicating confidence of edge existing in the true network

Weights are used to rank the edges, and different algorithms may have a different scale. The relative weights within one inferred network are therefore more meaningful than the absolute weight out of context.

NetworkInference.InferredNetwork — Type.

InferredNetwork type. Represents a weighted, fully connected network, where an edges's weight indicates the relative confidence of that edge existing in the true network.

Fields:

nodes: array of all the nodes, in an arbitrary order

edges: array of all the edges, in descending order of weight

NetworkInference.MINetworkInference

NetworkInference.NetworkInference

NetworkInference.Node — Type.

Node with metadata

Fields:

label: unique identifying label

binned_values: data values discretized into bins

number_of_bins: no. bins the data were discretized into

probabilities: probability distribution across the bins

NetworkInference.NodePair

NetworkInference.PIDCNetworkInference

NetworkInference.PUCNetworkInference

NetworkInference.apply_context

NetworkInference.eval

NetworkInference.get_adjacency_matrix — Function.

get_adjacency_matrix(inferred_network::InferredNetwork, threshold = 1.0; <keyword arguments>)

Gets an adjacency matrix given an InferredNetwork and a threshold.

Arguments:

inferred_network: network that was inferred

threshold=0.1: threshold above which to keep edges in the network

absolute=false: interpret threshold as an absolute confidence score

If absolute is false, threshold will be interpreted as the percentage of edges to keep.

NetworkInference.get_joint_probabilities

NetworkInference.get_mi_scores

NetworkInference.get_nodes — Function.

get_nodes(data_file_path::String; <keyword arguments>)

Gets an array of all Nodes from a data file. It is assumed that the first line of the file is headers (which are discarded) and the subsequent lines each represent one node, and are of the form:

Label datavalue1 datavalue2 ...

though a different delimiter may be specified.

Arguments:

data_file_path: path to the data file

delim=false: the file's delimiter. Leave as false if it is whitespace

discretizer="bayesian_blocks": algorithm for discretizing the data

estimator="maximum_likelihood": algorithm for estimating probabilities

number_of_bins=10: will be overwritten if using "bayesian_blocks"

The "maximum_likelihood" estimator is recommended for PUC and PIDC.

NetworkInference.get_puc

NetworkInference.get_puc_scores

NetworkInference.get_weight

NetworkInference.get_weights

NetworkInference.include

NetworkInference.infer_network — Function.

infer_network(data_file_path::String, inference::AbstractNetworkInference; <keyword arguments>)

Infers a network, given a data file and a network inference algorithm. It is assumed that the first line of the file is headers (which are discarded) and the subsequent lines each represent one node, and are of the form:

Label datavalue1 datavalue2 ...

though a different delimiter may be specified.

Arguments:

data_file_path: path to the data file

inference: network inference algorithm (e.g. PIDCNetworkInference())

delim=false: the file's delimiter. Leave as false if it is whitespace

discretizer="bayesian_blocks": algorithm for discretizing the data

estimator="maximum_likelihood": algorithm for estimating probabilities

number_of_bins=10: will be overwritten if using "bayesian_blocks"

base=2: base for the information measures

out_file_path="": path to output file. If empty, will not write a file

The "maximum_likelihood" estimator is recommended for PUC and PIDC.

NetworkInference.read_network_file — Function.

read_network_file(file_path::AbstractString)

Reads a network file and creates an InferredNetwork type. Assumes that the input is such that each line contains an edge and each edge is written in both directions with the same weight:

...

LabelX LabelY WeightXY

LabelY LabelX WeightXY

...

NetworkInference.write_network_file — Function.

write_network_file(file_path::String, inferred_network::InferredNetwork)

Writes a network file from an InferredNetwork type. Each line of the file will contain an edge, and since networks are assumed undirected, each edge will be written in both directions with the same weight:

...

LabelX LabelY WeightXY

LabelY LabelX WeightXY

...

Arguments:

file_path: path to the output file

inferred_network: network that was inferred