Docstrings

MambaModels.+

MambaModels.-

MambaModels.@animate

MambaModels.@colorant_str

MambaModels.@csv2_str

MambaModels.@csv_str

MambaModels.@df

MambaModels.@gif

MambaModels.@layout

MambaModels.@recipe

MambaModels.@series

MambaModels.@shorthands

MambaModels.@tsv_str

MambaModels.@userplot

MambaModels.@wsv_str

MambaModels.ABC

MambaModels.ABGR

MambaModels.ADIN99

MambaModels.ADIN99d

MambaModels.ADIN99o

MambaModels.AGray

MambaModels.AGray32

MambaModels.AHSI

MambaModels.AHSL

MambaModels.AHSV

MambaModels.ALCHab

MambaModels.ALCHuv

MambaModels.ALMS

MambaModels.ALab

MambaModels.ALuv

MambaModels.AMM

MambaModels.AMMVariate

MambaModels.AMWG

MambaModels.AMWGVariate

MambaModels.ARGB

MambaModels.ARGB32

MambaModels.AXYZ

MambaModels.AYCbCr

MambaModels.AYIQ

MambaModels.AbstractBackend

MambaModels.AbstractCategoricalArray

MambaModels.AbstractCategoricalMatrix

MambaModels.AbstractCategoricalVector

MambaModels.AbstractChains

MambaModels.AbstractDataFrame

MambaModels.AbstractDependent

MambaModels.AbstractGray

MambaModels.AbstractHistogram

MambaModels.AbstractLayout

MambaModels.AbstractLogical

MambaModels.AbstractMissingCategoricalArray

MambaModels.AbstractMissingCategoricalMatrix

MambaModels.AbstractMissingCategoricalVector

MambaModels.AbstractMixtureModel

MambaModels.AbstractMvNormal

MambaModels.AbstractPlot

MambaModels.AbstractRGB

MambaModels.AbstractSerializer

MambaModels.AbstractStochastic

MambaModels.AbstractVariate

MambaModels.AbstractWeights

MambaModels.AlphaColor

MambaModels.AnalyticWeights

MambaModels.Animation

MambaModels.Arcsine

MambaModels.ArrayLogical

MambaModels.ArrayStochastic

MambaModels.ArrayVariate

MambaModels.AxyY

MambaModels.BDiagNormal

MambaModels.BGR

MambaModels.BGRA

MambaModels.BHMC

MambaModels.BHMCVariate

MambaModels.BIA

MambaModels.BIAVariate

MambaModels.BMC3

MambaModels.BMC3Variate

MambaModels.BMG

MambaModels.BMGVariate

MambaModels.Bernoulli

MambaModels.Beta

MambaModels.BetaBinomial

MambaModels.BetaPrime

MambaModels.BezierCurve

MambaModels.Binomial

MambaModels.Biweight

MambaModels.CIE1931JV_CMF

MambaModels.CIE1931J_CMF

MambaModels.CIE1931_CMF

MambaModels.CIE1964_CMF

MambaModels.CSV

MambaModels.Categorical

MambaModels.CategoricalArray

MambaModels.CategoricalArrays

MambaModels.CategoricalMatrix

MambaModels.CategoricalPool

MambaModels.CategoricalString

MambaModels.CategoricalValue

MambaModels.CategoricalVector

MambaModels.Cauchy

MambaModels.Chains

MambaModels.Chi

MambaModels.Chisq

MambaModels.CoefTable

MambaModels.Color

MambaModels.Color3

MambaModels.ColorAlpha

MambaModels.ColorGradient

MambaModels.ColorTypes

MambaModels.Colorant

MambaModels.ColorantNormed

MambaModels.Colors

MambaModels.Continuous

MambaModels.ContinuousDistribution

MambaModels.ContinuousMatrixDistribution

MambaModels.ContinuousMultivariateDistribution

MambaModels.ContinuousUnivariateDistribution

MambaModels.ConvergenceException

MambaModels.Cosine

MambaModels.DE_2000

MambaModels.DE_94

MambaModels.DE_AB

MambaModels.DE_BFD

MambaModels.DE_CMC

MambaModels.DE_DIN99

MambaModels.DE_DIN99d

MambaModels.DE_DIN99o

MambaModels.DE_JPC79

MambaModels.DGS

MambaModels.DGSVariate

MambaModels.DIN99

MambaModels.DIN99A

MambaModels.DIN99d

MambaModels.DIN99dA

MambaModels.DIN99o

MambaModels.DIN99oA

MambaModels.DataFrame

MambaModels.DataFrameRow

MambaModels.DataFrames

MambaModels.DelimitedFiles

MambaModels.DiagNormal

MambaModels.DiagNormalCanon

MambaModels.Dirichlet

MambaModels.DirichletMultinomial

MambaModels.Discrete

MambaModels.DiscreteDistribution

MambaModels.DiscreteMatrixDistribution

MambaModels.DiscreteMultivariateDistribution

MambaModels.DiscreteUniform

MambaModels.DiscreteUnivariateDistribution

MambaModels.DiscreteVariate

MambaModels.Distribution

MambaModels.Distributions

MambaModels.DoubleExponential

MambaModels.ECDF

MambaModels.EdgeworthMean

MambaModels.EdgeworthSum

MambaModels.EdgeworthZ

MambaModels.EmpiricalUnivariateDistribution

MambaModels.Epanechnikov

MambaModels.Erlang

MambaModels.Estimator

MambaModels.Exponential

MambaModels.FDist

MambaModels.FisherNoncentralHypergeometric

MambaModels.Flat

MambaModels.Formatted

MambaModels.Fractional

MambaModels.Frechet

MambaModels.FrequencyWeights

MambaModels.FullNormal

MambaModels.FullNormalCanon

MambaModels.GR

MambaModels.Gamma

MambaModels.GeneralizedExtremeValue

MambaModels.GeneralizedPareto

MambaModels.Geometric

MambaModels.Gray

MambaModels.Gray24

MambaModels.GrayA

MambaModels.GroupApplied

MambaModels.GroupedDataFrame

MambaModels.Gumbel

MambaModels.HMC

MambaModels.HMCVariate

MambaModels.HSB

MambaModels.HSI

MambaModels.HSIA

MambaModels.HSL

MambaModels.HSLA

MambaModels.HSV

MambaModels.HSVA

MambaModels.Histogram

MambaModels.Hypergeometric

MambaModels.InverseGamma

MambaModels.InverseGaussian

MambaModels.InverseWishart

MambaModels.IsoNormal

MambaModels.IsoNormalCanon

MambaModels.KSDist

MambaModels.KSOneSided

MambaModels.KW

MambaModels.Kolmogorov

MambaModels.L1dist

MambaModels.L2dist

MambaModels.LCHab

MambaModels.LCHabA

MambaModels.LCHuv

MambaModels.LCHuvA

MambaModels.LMS

MambaModels.LMSA

MambaModels.Lab

MambaModels.LabA

MambaModels.Laplace

MambaModels.LevelsException

MambaModels.Levy

MambaModels.Linfdist

MambaModels.LocationScale

MambaModels.LogNormal

MambaModels.Logical

MambaModels.Logistic

MambaModels.Luv

MambaModels.LuvA

MambaModels.MALA

MambaModels.MALAVariate

MambaModels.MCMCChain

MambaModels.MISS

MambaModels.MLEstimator

MambaModels.MSC

MambaModels.Mamba

MambaModels.MambaModels

MambaModels.MatrixDistribution

MambaModels.MatrixVariate

MambaModels.Matrixvariate

MambaModels.Missing

MambaModels.MissingCategoricalArray

MambaModels.MissingCategoricalMatrix

MambaModels.MissingCategoricalVector

MambaModels.MissingException

MambaModels.Missings

MambaModels.MixtureModel

MambaModels.Model

MambaModels.ModelChains

MambaModels.Multinomial

MambaModels.Multivariate

MambaModels.MultivariateDistribution

MambaModels.MultivariateMixture

MambaModels.MultivariateNormal

MambaModels.MvLogNormal

MambaModels.MvNormal

MambaModels.MvNormalCanon

MambaModels.MvNormalKnownCov

MambaModels.MvTDist

MambaModels.NUTS

MambaModels.NUTSVariate

MambaModels.NegativeBinomial

MambaModels.NonMatrixDistribution

MambaModels.NoncentralBeta

MambaModels.NoncentralChisq

MambaModels.NoncentralF

MambaModels.NoncentralHypergeometric

MambaModels.NoncentralT

MambaModels.Normal

MambaModels.NormalCanon

MambaModels.NormalInverseGaussian

MambaModels.OHLC

MambaModels.OrderedLevelsException

MambaModels.Pareto

MambaModels.PlotThemes

MambaModels.PlotUtils

MambaModels.Plots

MambaModels.Poisson

MambaModels.PoissonBinomial

MambaModels.ProbabilityWeights

MambaModels.QQPair

MambaModels.RDatasets

MambaModels.RGB

MambaModels.RGB1

MambaModels.RGB24

MambaModels.RGB4

MambaModels.RGBA

MambaModels.RWM

MambaModels.RWMVariate

MambaModels.Rayleigh

MambaModels.RealInterval

MambaModels.RecipeData

MambaModels.RecipesBase

MambaModels.RegressionModel

MambaModels.Sampleable

MambaModels.Sampler

MambaModels.SamplerTune

MambaModels.SamplerVariate

MambaModels.ScalarLogical

MambaModels.ScalarStochastic

MambaModels.ScalarVariate

MambaModels.ScriptEntry

MambaModels.Segments

MambaModels.Semicircle

MambaModels.Serialization

MambaModels.Serializer

MambaModels.Shape

MambaModels.Skellam

MambaModels.Slice

MambaModels.SliceMultivariate

MambaModels.SliceSimplex

MambaModels.SliceSimplexVariate

MambaModels.SliceUnivariate

MambaModels.StatisticalModel

MambaModels.StatisticalRethinking

MambaModels.StatsBase

MambaModels.StatsFuns

MambaModels.StatsPlots

MambaModels.Stochastic

MambaModels.SubDataFrame

MambaModels.SufficientStats

MambaModels.Surface

MambaModels.SymTriangularDist

MambaModels.SymUniform

MambaModels.TDist

MambaModels.Transparent3

MambaModels.TransparentColor

MambaModels.TransparentGray

MambaModels.TransparentRGB

MambaModels.TriangularDist

MambaModels.Triweight

MambaModels.Truncated

MambaModels.TruncatedNormal

MambaModels.Uniform

MambaModels.Univariate

MambaModels.UnivariateDistribution

MambaModels.UnivariateGMM

MambaModels.UnivariateMixture

MambaModels.ValueSupport

MambaModels.VariateForm

MambaModels.VectorVariate

MambaModels.VonMises

MambaModels.VonMisesFisher

MambaModels.WalleniusNoncentralHypergeometric

MambaModels.Weibull

MambaModels.Weights

MambaModels.Wishart

MambaModels.XYZ

MambaModels.XYZA

MambaModels.YCbCr

MambaModels.YCbCrA

MambaModels.YIQ

MambaModels.YIQA

MambaModels.ZeroMeanDiagNormal

MambaModels.ZeroMeanDiagNormalCanon

MambaModels.ZeroMeanFullNormal

MambaModels.ZeroMeanFullNormalCanon

MambaModels.ZeroMeanIsoNormal

MambaModels.ZeroMeanIsoNormalCanon

MambaModels.adapted_grid

MambaModels.add_backend

MambaModels.add_theme

MambaModels.addcounts!

MambaModels.adjr2

MambaModels.adjr²

MambaModels.aggregate

MambaModels.aic

MambaModels.aicc

MambaModels.aliases

MambaModels.allowmissing

MambaModels.allowmissing!

MambaModels.alpha

MambaModels.alphacolor

MambaModels.andrewsplot

MambaModels.andrewsplot!

MambaModels.animate

MambaModels.annotate!

MambaModels.arrow

MambaModels.attr!

MambaModels.autcorplot

MambaModels.autocor

MambaModels.autocor!

MambaModels.autocorplot

MambaModels.autocorplot!

MambaModels.autocov

MambaModels.autocov!

MambaModels.aweights

MambaModels.backend

MambaModels.backend_name

MambaModels.backend_object

MambaModels.backends

MambaModels.bar

MambaModels.bar!

MambaModels.barh

MambaModels.barh!

MambaModels.barhist

MambaModels.barhist!

MambaModels.base_color_type

MambaModels.base_colorant_type

MambaModels.bbox

MambaModels.betaccdf

MambaModels.betacdf

MambaModels.betainvccdf

MambaModels.betainvcdf

MambaModels.betainvlogccdf

MambaModels.betainvlogcdf

MambaModels.betalogccdf

MambaModels.betalogcdf

MambaModels.betalogpdf

MambaModels.betapdf

MambaModels.bic

MambaModels.binaryentropy

MambaModels.binomccdf

MambaModels.binomcdf

MambaModels.binominvccdf

MambaModels.binominvcdf

MambaModels.binominvlogccdf

MambaModels.binominvlogcdf

MambaModels.binomlogccdf

MambaModels.binomlogcdf

MambaModels.binomlogpdf

MambaModels.binompdf

MambaModels.blue

MambaModels.boxplot

MambaModels.boxplot!

MambaModels.brush

MambaModels.by

MambaModels.canonform

MambaModels.categorical

MambaModels.categorical!

MambaModels.ccdf

MambaModels.ccolor

MambaModels.cdensity

MambaModels.cdensity!

MambaModels.cdf

MambaModels.center

MambaModels.cf

MambaModels.cgf

MambaModels.cgrad

MambaModels.cgradients

MambaModels.changerate

MambaModels.chisqccdf

MambaModels.chisqcdf

MambaModels.chisqinvccdf

MambaModels.chisqinvcdf

MambaModels.chisqinvlogccdf

MambaModels.chisqinvlogcdf

MambaModels.chisqlogccdf

MambaModels.chisqlogcdf

MambaModels.chisqlogpdf

MambaModels.chisqpdf

MambaModels.cie_color_match

MambaModels.circvar

MambaModels.clibraries

MambaModels.clibrary

MambaModels.closeall

MambaModels.cm

MambaModels.coalesce

MambaModels.coef

MambaModels.coefnames

MambaModels.coeftable

MambaModels.color

MambaModels.color_type

MambaModels.coloralpha

MambaModels.colordiff

MambaModels.colormap

MambaModels.colormatch

MambaModels.colwise

MambaModels.combine

MambaModels.comp1

MambaModels.comp2

MambaModels.comp3

MambaModels.compact

MambaModels.competerank

MambaModels.complete_cases

MambaModels.complete_cases!

MambaModels.completecases

MambaModels.component

MambaModels.components

MambaModels.componentwise_logpdf

MambaModels.componentwise_pdf

MambaModels.compress

MambaModels.concentration

MambaModels.confint

MambaModels.contour

MambaModels.contour!

MambaModels.contour3d

MambaModels.contour3d!

MambaModels.contourf

MambaModels.contourf!

MambaModels.convert

MambaModels.coords

MambaModels.cor

MambaModels.cor2cov

MambaModels.corkendall

MambaModels.corner

MambaModels.corner!

MambaModels.cornerplot

MambaModels.cornerplot!

MambaModels.corrplot

MambaModels.corrplot!

MambaModels.corspearman

MambaModels.counteq

MambaModels.countmap

MambaModels.countne

MambaModels.counts

MambaModels.cov

MambaModels.cov2cor

MambaModels.cquantile

MambaModels.crosscor

MambaModels.crosscor!

MambaModels.crosscov

MambaModels.crosscov!

MambaModels.crossentropy

MambaModels.cumulant

MambaModels.current

MambaModels.curve_points

MambaModels.curves

MambaModels.curves!

MambaModels.cut

MambaModels.cvec

MambaModels.dataset

MambaModels.dataviewer

MambaModels.decompress

MambaModels.default

MambaModels.default_cgrad

MambaModels.delete!

MambaModels.deletecols!

MambaModels.deleterows!

MambaModels.denserank

MambaModels.density

MambaModels.density!

MambaModels.densityplot

MambaModels.describe

MambaModels.deserialize

MambaModels.deuteranopic

MambaModels.deviance

MambaModels.dic

MambaModels.dim

MambaModels.disallowmissing

MambaModels.disallowmissing!

MambaModels.discretediag

MambaModels.distinguishable_colors

MambaModels.diverging_palette

MambaModels.dof

MambaModels.dof_residual

MambaModels.draw

MambaModels.droplevels!

MambaModels.dropmissing

MambaModels.dropmissing!

MambaModels.ea_histogram

MambaModels.ea_histogram!

MambaModels.ecdf

MambaModels.eltypes

MambaModels.entropy

MambaModels.estimate

MambaModels.eval

MambaModels.expected_logdet

MambaModels.failprob

MambaModels.fdistccdf

MambaModels.fdistcdf

MambaModels.fdistinvccdf

MambaModels.fdistinvcdf

MambaModels.fdistinvlogccdf

MambaModels.fdistinvlogcdf

MambaModels.fdistlogccdf

MambaModels.fdistlogcdf

MambaModels.fdistlogpdf

MambaModels.fdistpdf

MambaModels.findat

MambaModels.fit

MambaModels.fit!

MambaModels.fit_map

MambaModels.fit_map!

MambaModels.fit_mle

MambaModels.fit_mle!

MambaModels.fitted

MambaModels.font

MambaModels.fourinvπ

MambaModels.fourπ

MambaModels.frame

MambaModels.freecumulant

MambaModels.fweights

MambaModels.gammaccdf

MambaModels.gammacdf

MambaModels.gammainvccdf

MambaModels.gammainvcdf

MambaModels.gammainvlogccdf

MambaModels.gammainvlogcdf

MambaModels.gammalogccdf

MambaModels.gammalogcdf

MambaModels.gammalogpdf

MambaModels.gammapdf

MambaModels.gamutmax

MambaModels.gamutmin

MambaModels.gelmandiag

MambaModels.generate

MambaModels.generate_m — Function.

generate

Utility function to generate all notebooks and chapters from scripts in the scripts directory.

Method

generate_m(sd = script_dict_m)

Required arguments

None, all notebooks/.. and chapters/.. files are regenerated.

generate

Generate notebooks and scripts in a single chapter.

Method

generate_m(chapter::AbstractString)

Required arguments

Generate notebooks and scripts in a single chapter, e.g. generate_m("04")

generate

Generate a single notebook and script

Method

generate_m(chapter::AbstractString, file::AbstractString)

Required arguments

Generate notebook and script file in chapter, e.g. generatem("04", "m4.1d.jl") or generatem("04/m4.1d.jl")

MambaModels.genmean

MambaModels.genvar

MambaModels.geomean

MambaModels.get_color_palette

MambaModels.getindex

MambaModels.gettune

MambaModels.gewekediag

MambaModels.gif

MambaModels.gkldiv

MambaModels.gr

MambaModels.gradlogpdf

MambaModels.gradlogpdf!

MambaModels.graph

MambaModels.graph2dot

MambaModels.gray

MambaModels.green

MambaModels.grid

MambaModels.groupby

MambaModels.groupedbar

MambaModels.groupedbar!

MambaModels.groupindices

MambaModels.groupvars

MambaModels.gui

MambaModels.halfπ

MambaModels.harmmean

MambaModels.hasfinitesupport

MambaModels.hdf5

MambaModels.head

MambaModels.heatmap

MambaModels.heatmap!

MambaModels.heideldiag

MambaModels.hex

MambaModels.hexbin

MambaModels.hexbin!

MambaModels.hist

MambaModels.histogram

MambaModels.histogram!

MambaModels.histogram2d

MambaModels.histogram2d!

MambaModels.histogramplot

MambaModels.hline

MambaModels.hline!

MambaModels.hpd

MambaModels.hspan

MambaModels.hspan!

MambaModels.hyperccdf

MambaModels.hypercdf

MambaModels.hyperinvccdf

MambaModels.hyperinvcdf

MambaModels.hyperinvlogccdf

MambaModels.hyperinvlogcdf

MambaModels.hyperlogccdf

MambaModels.hyperlogcdf

MambaModels.hyperlogpdf

MambaModels.hyperpdf

MambaModels.inch

MambaModels.include

MambaModels.indexmap

MambaModels.indicatormat

MambaModels.informationmatrix

MambaModels.inline

MambaModels.insert!

MambaModels.insertcols!

MambaModels.inspectdr

MambaModels.insupport

MambaModels.inv2π

MambaModels.inv4π

MambaModels.invcov

MambaModels.inverse_rle

MambaModels.invisible

MambaModels.invlink

MambaModels.invlogccdf

MambaModels.invlogcdf

MambaModels.invlogit

MambaModels.invscale

MambaModels.invsoftplus

MambaModels.invsqrt2

MambaModels.invsqrt2π

MambaModels.invπ

MambaModels.iqr

MambaModels.isbounded

MambaModels.isdark

MambaModels.isfitted

MambaModels.isleptokurtic

MambaModels.islinear

MambaModels.islowerbounded

MambaModels.ismesokurtic

MambaModels.ismissing

MambaModels.isordered

MambaModels.isplatykurtic

MambaModels.isprobvec

MambaModels.isupperbounded

MambaModels.iter_segments

MambaModels.kde

MambaModels.keys

MambaModels.kldivergence

MambaModels.kurtosis

MambaModels.length

MambaModels.levels

MambaModels.levels!

MambaModels.levelsmap

MambaModels.leverage

MambaModels.link

MambaModels.location

MambaModels.location!

MambaModels.log1mexp

MambaModels.log1pexp

MambaModels.log1pmx

MambaModels.log1psq

MambaModels.log2mexp

MambaModels.log2π

MambaModels.log4π

MambaModels.logaddexp

MambaModels.logccdf

MambaModels.logcdf

MambaModels.logdetcov

MambaModels.logexpm1

MambaModels.loghalf

MambaModels.logistic

MambaModels.logit

MambaModels.loglikelihood

MambaModels.logmvgamma

MambaModels.logmxp1

MambaModels.logpdf

MambaModels.logpdf!

MambaModels.logsumexp

MambaModels.logtwo

MambaModels.logπ

MambaModels.lstirling_asym

MambaModels.mad

MambaModels.mad!

MambaModels.map

MambaModels.mapc

MambaModels.mapcols

MambaModels.mapreducec

MambaModels.marginalhist

MambaModels.marginalhist!

MambaModels.maxad

MambaModels.mcmc

MambaModels.mcse

MambaModels.mean

MambaModels.mean_and_cov

MambaModels.mean_and_std

MambaModels.mean_and_var

MambaModels.meanad

MambaModels.meandir

MambaModels.meanform

MambaModels.meanlogx

MambaModels.meanplot

MambaModels.meanplot!

MambaModels.meanresponse

MambaModels.median

MambaModels.melt

MambaModels.meltdf

MambaModels.merge!

MambaModels.mgf

MambaModels.middle

MambaModels.midpoints

MambaModels.missing

MambaModels.missings

MambaModels.mixeddensity

MambaModels.mixeddensity!

MambaModels.mixeddensityplot

MambaModels.mm

MambaModels.mode

MambaModels.model_response

MambaModels.modelmatrix

MambaModels.modes

MambaModels.moment

MambaModels.mov

MambaModels.mp4

MambaModels.msd

MambaModels.mss

MambaModels.names!

MambaModels.nbetaccdf

MambaModels.nbetacdf

MambaModels.nbetainvccdf

MambaModels.nbetainvcdf

MambaModels.nbetainvlogccdf

MambaModels.nbetainvlogcdf

MambaModels.nbetalogccdf

MambaModels.nbetalogcdf

MambaModels.nbetalogpdf

MambaModels.nbetapdf

MambaModels.nbinomccdf

MambaModels.nbinomcdf

MambaModels.nbinominvccdf

MambaModels.nbinominvcdf

MambaModels.nbinominvlogccdf

MambaModels.nbinominvlogcdf

MambaModels.nbinomlogccdf

MambaModels.nbinomlogcdf

MambaModels.nbinomlogpdf

MambaModels.nbinompdf

MambaModels.ncategories

MambaModels.nchisqccdf

MambaModels.nchisqcdf

MambaModels.nchisqinvccdf

MambaModels.nchisqinvcdf

MambaModels.nchisqinvlogccdf

MambaModels.nchisqinvlogcdf

MambaModels.nchisqlogccdf

MambaModels.nchisqlogcdf

MambaModels.nchisqlogpdf

MambaModels.nchisqpdf

MambaModels.ncol

MambaModels.ncomponents

MambaModels.nfdistccdf

MambaModels.nfdistcdf

MambaModels.nfdistinvccdf

MambaModels.nfdistinvcdf

MambaModels.nfdistinvlogccdf

MambaModels.nfdistinvlogcdf

MambaModels.nfdistlogccdf

MambaModels.nfdistlogcdf

MambaModels.nfdistlogpdf

MambaModels.nfdistpdf

MambaModels.nobs

MambaModels.nonunique

MambaModels.norepeats

MambaModels.normccdf

MambaModels.normcdf

MambaModels.norminvccdf

MambaModels.norminvcdf

MambaModels.norminvlogccdf

MambaModels.norminvlogcdf

MambaModels.normlogccdf

MambaModels.normlogcdf

MambaModels.normlogpdf

MambaModels.normpdf

MambaModels.nquantile

MambaModels.nrow

MambaModels.nsamples

MambaModels.ntdistccdf

MambaModels.ntdistcdf

MambaModels.ntdistinvccdf

MambaModels.ntdistinvcdf

MambaModels.ntdistinvlogccdf

MambaModels.ntdistinvlogcdf

MambaModels.ntdistlogccdf

MambaModels.ntdistlogcdf

MambaModels.ntdistlogpdf

MambaModels.ntdistpdf

MambaModels.ntrials

MambaModels.nullable!

MambaModels.nulldeviance

MambaModels.nullloglikelihood

MambaModels.ohlc

MambaModels.ohlc!

MambaModels.optimize_datetime_ticks

MambaModels.optimize_ticks

MambaModels.order

MambaModels.ordered

MambaModels.ordered!

MambaModels.ordinalrank

MambaModels.pacf

MambaModels.pacf!

MambaModels.palette

MambaModels.params

MambaModels.params!

MambaModels.partialcor

MambaModels.partype

MambaModels.passmissing

MambaModels.path3d

MambaModels.path3d!

MambaModels.pdf

MambaModels.percentile

MambaModels.permutecols!

MambaModels.pgfplots

MambaModels.pie

MambaModels.pie!

MambaModels.plot

MambaModels.plot!

MambaModels.plot3d

MambaModels.plot3d!

MambaModels.plot_color

MambaModels.plotarea

MambaModels.plotattr

MambaModels.plotly

MambaModels.plotlyjs

MambaModels.plots_heatmap

MambaModels.plots_heatmap!

MambaModels.png

MambaModels.poisccdf

MambaModels.poiscdf

MambaModels.poisinvccdf

MambaModels.poisinvcdf

MambaModels.poisinvlogccdf

MambaModels.poisinvlogcdf

MambaModels.poislogccdf

MambaModels.poislogcdf

MambaModels.poislogpdf

MambaModels.poispdf

MambaModels.pool

MambaModels.pool!

MambaModels.portfoliocomposition

MambaModels.portfoliocomposition!

MambaModels.predict

MambaModels.predict!

MambaModels.probs

MambaModels.probval

MambaModels.proportionmap

MambaModels.proportions

MambaModels.protanopic

MambaModels.psnr

MambaModels.pt

MambaModels.pweights

MambaModels.px

MambaModels.pyplot

MambaModels.qqbuild

MambaModels.qqnorm

MambaModels.qqnorm!

MambaModels.qqplot

MambaModels.qqplot!

MambaModels.quantile

MambaModels.quartπ

MambaModels.quiver

MambaModels.quiver!

MambaModels.r2

MambaModels.rafterydiag

MambaModels.rand

MambaModels.rate

MambaModels.readcoda

MambaModels.readdlm

MambaModels.readtable

MambaModels.recode

MambaModels.recode!

MambaModels.red

MambaModels.reducec

MambaModels.rel_path

MambaModels.rel_path_m — Function.

relpathm

Relative path using the MambaModels src/ directory.

Example to get access to the data subdirectory

rel_path_m("..", "data")

MambaModels.relist

MambaModels.relist!

MambaModels.rename

MambaModels.rename!

MambaModels.renyientropy

MambaModels.residuals

MambaModels.response

MambaModels.rgb_string

MambaModels.rgba_string

MambaModels.rle

MambaModels.rmsd

MambaModels.rotate

MambaModels.rotate!

MambaModels.rss

MambaModels.r²

MambaModels.sample

MambaModels.sample!

MambaModels.samplepair

MambaModels.sampler

MambaModels.savefig

MambaModels.scale

MambaModels.scale!

MambaModels.scatter

MambaModels.scatter!

MambaModels.scatter3d

MambaModels.scatter3d!

MambaModels.scatterhist

MambaModels.scatterhist!

MambaModels.scattermat

MambaModels.score

MambaModels.script_dict

MambaModels.script_dict_m

MambaModels.scriptentry

MambaModels.sem

MambaModels.sequential_palette

MambaModels.serialize

MambaModels.set_theme

MambaModels.setindex!

MambaModels.setinits!

MambaModels.setinputs!

MambaModels.setmonitor!

MambaModels.setsamplers!

MambaModels.shape

MambaModels.shape_coords

MambaModels.show

MambaModels.showall

MambaModels.showcols

MambaModels.showgradient

MambaModels.showgradient!

MambaModels.showlibrary

MambaModels.showlibrary!

MambaModels.showtheme

MambaModels.showtheme!

MambaModels.skewness

MambaModels.softmax

MambaModels.softmax!

MambaModels.softplus

MambaModels.span

MambaModels.spy

MambaModels.spy!

MambaModels.sqL2dist

MambaModels.sqmahal

MambaModels.sqmahal!

MambaModels.sqrt2

MambaModels.sqrt2π

MambaModels.sqrt3

MambaModels.sqrt4π

MambaModels.sqrthalfπ

MambaModels.sqrtπ

MambaModels.src_path_m

MambaModels.srdistccdf

MambaModels.srdistcdf

MambaModels.srdistinvccdf

MambaModels.srdistinvcdf

MambaModels.srdistinvlogccdf

MambaModels.srdistinvlogcdf

MambaModels.srdistlogccdf

MambaModels.srdistlogcdf

MambaModels.stack

MambaModels.stackdf

MambaModels.std

MambaModels.stderror

MambaModels.stdlogx

MambaModels.stephist

MambaModels.stephist!

MambaModels.sticks

MambaModels.sticks!

MambaModels.stroke

MambaModels.sub

MambaModels.succprob

MambaModels.suffstats

MambaModels.summarystats

MambaModels.support

MambaModels.surface

MambaModels.surface!

MambaModels.tail

MambaModels.tdistccdf

MambaModels.tdistcdf

MambaModels.tdistinvccdf

MambaModels.tdistinvcdf

MambaModels.tdistinvlogccdf

MambaModels.tdistinvlogcdf

MambaModels.tdistlogccdf

MambaModels.tdistlogcdf

MambaModels.tdistlogpdf

MambaModels.tdistpdf

MambaModels.test_distr

MambaModels.test_examples

MambaModels.test_samples

MambaModels.testhist

MambaModels.testhist!

MambaModels.text

MambaModels.theme

MambaModels.tiedrank

MambaModels.title!

MambaModels.totalvar

MambaModels.traceplot

MambaModels.traceplot!

MambaModels.transform

MambaModels.translate

MambaModels.translate!

MambaModels.trim

MambaModels.trim!

MambaModels.trimvar

MambaModels.tritanopic

MambaModels.twinx

MambaModels.twoinvπ

MambaModels.twoπ

MambaModels.uncompact

MambaModels.unicodeplots

MambaModels.unique!

MambaModels.unlist

MambaModels.unstack

MambaModels.update!

MambaModels.values

MambaModels.var

MambaModels.variation

MambaModels.varlogx

MambaModels.vcat

MambaModels.vcov

MambaModels.violin

MambaModels.violin!

MambaModels.vline

MambaModels.vline!

MambaModels.vspan

MambaModels.vspan!

MambaModels.weighted_color_mean

MambaModels.weights

MambaModels.whitebalance

MambaModels.winsor

MambaModels.winsor!

MambaModels.wireframe

MambaModels.wireframe!

MambaModels.with

MambaModels.wmean

MambaModels.wmean!

MambaModels.wmedian

MambaModels.wquantile

MambaModels.wrap

MambaModels.writedlm

MambaModels.writetable

MambaModels.wsample

MambaModels.wsample!

MambaModels.wsum

MambaModels.wsum!

MambaModels.xaxis!

MambaModels.xflip!

MambaModels.xgrid!

MambaModels.xlabel!

MambaModels.xlims

MambaModels.xlims!

MambaModels.xlogx

MambaModels.xlogy

MambaModels.xticks!

MambaModels.xyY

MambaModels.xyYA

MambaModels.yaxis!

MambaModels.yflip!

MambaModels.ygrid!

MambaModels.ylabel!

MambaModels.ylims

MambaModels.ylims!

MambaModels.yticks!

MambaModels.zlims

MambaModels.zlims!

MambaModels.zscore

MambaModels.zscore!