Docstrings

OptimizationProblems.AMPGO02 — Function.

Univariate multimodal minimization problem AMPGO02

OptimizationProblems.AMPGO03 — Function.

Univariate multimodal minimization problem AMPGO03

OptimizationProblems.AMPGO04 — Function.

Univariate multimodal minimization problem AMPGO04

OptimizationProblems.AMPGO05 — Function.

Univariate multimodal minimization problem AMPGO05

OptimizationProblems.AMPGO06 — Function.

Univariate multimodal minimization problem AMPGO06

OptimizationProblems.AMPGO07 — Function.

Univariate multimodal minimization problem AMPGO07

OptimizationProblems.AMPGO08 — Function.

Univariate multimodal minimization problem AMPGO08

OptimizationProblems.AMPGO09 — Function.

Univariate multimodal minimization problem AMPGO09

OptimizationProblems.AMPGO10 — Function.

Univariate multimodal minimization problem AMPGO10

OptimizationProblems.AMPGO11 — Function.

Univariate multimodal minimization problem AMPGO11

OptimizationProblems.AMPGO12 — Function.

Univariate multimodal minimization problem AMPGO12

OptimizationProblems.AMPGO13 — Function.

Univariate multimodal minimization problem AMPGO13

OptimizationProblems.AMPGO14 — Function.

Univariate multimodal minimization problem AMPGO14

OptimizationProblems.AMPGO15 — Function.

Univariate multimodal minimization problem AMPGO15

OptimizationProblems.AMPGO18 — Function.

Univariate multimodal minimization problem AMPGO18

OptimizationProblems.AMPGO20 — Function.

Univariate multimodal minimization problem AMPGO20

OptimizationProblems.AMPGO21 — Function.

Univariate multimodal minimization problem AMPGO21

OptimizationProblems.AMPGO22 — Function.

Univariate multimodal minimization problem AMPGO22

OptimizationProblems.Dus2_1 — Function.

Univariate unimodal minimization problem Dus2_1

OptimizationProblems.Dus2_3 — Function.

Univariate unimodal minimization problem Dus2_3

OptimizationProblems.Dus2_9 — Function.

Univariate multimodal minimization problem Dus2_9

OptimizationProblems.Duscube — Function.

Univariate multimodal minimization problem Duscube

OptimizationProblems.NZF1

OptimizationProblems.OptimizationProblems

OptimizationProblems.Shpak1 — Function.

Univariate multimodal minimization problem Shpak1

OptimizationProblems.Shpak2 — Function.

Univariate multimodal minimization problem Shpak2

OptimizationProblems.Shpak3 — Function.

Univariate multimodal minimization problem Shpak3

OptimizationProblems.Shpak4 — Function.

Univariate multimodal minimization problem Shpak4

OptimizationProblems.Shpak5 — Function.

Univariate multimodal minimization problem Shpak5

OptimizationProblems.Shpak6 — Function.

Univariate multimodal minimization problem Shpak6

OptimizationProblems.arglina — Function.

Linear function with n parameters and m observations - full rank

OptimizationProblems.arglinb — Function.

Linear function with n parameters and m observations - rank 1

OptimizationProblems.arglinc — Function.

Linear function with n parameters and m observations - rank 1, zero columns and rows

OptimizationProblems.arwhead — Function.

Arrow head model in size n

OptimizationProblems.bdqrtic — Function.

Banded quartic model in size n

OptimizationProblems.beale — Function.

Beale Model

OptimizationProblems.brownbs — Function.

Brownbs Model

OptimizationProblems.brownden — Function.

Brown and Dennis function

OptimizationProblems.broydn7d — Function.

Broyden 7-diagonal model in size n

OptimizationProblems.brybnd — Function.

Broyden banded model in size n

OptimizationProblems.bt1 — Function.

BT1 Model

OptimizationProblems.chainwoo — Function.

The chained Woods function in size n, a variant on the Woods function

OptimizationProblems.chnrosnb_mod

OptimizationProblems.cliff — Function.

The 'cliff problem' in 2 variables

OptimizationProblems.clplatea — Function.

The clamped plate problem (Strang, Nocedal, Dax).

OptimizationProblems.clplateb — Function.

The clamped plate problem (Strang, Nocedal, Dax).

OptimizationProblems.clplatec — Function.

The clamped plate problem (Strang, Nocedal, Dax).

OptimizationProblems.cosine — Function.

The cosine function in size n

OptimizationProblems.cragglvy — Function.

The extented Cragg and Levy function in size n

OptimizationProblems.curly — Function.

Curly function in size n with semi-bandwidth b

OptimizationProblems.curly10 — Function.

Curly function in size n with semi-bandwidth 10

OptimizationProblems.curly20 — Function.

Curly function in size n with semi-bandwidth 20

OptimizationProblems.curly30 — Function.

Curly function in size n with semi-bandwidth 30

OptimizationProblems.dixmaane — Function.

Dixon-Maany function in size n (version E by default)

OptimizationProblems.dixmaanf — Function.

Dixon-Maany function in size n (version F)

OptimizationProblems.dixmaang — Function.

Dixon-Maany function in size n (version G)

OptimizationProblems.dixmaanh — Function.

Dixon-Maany function in size n (version H)

OptimizationProblems.dixmaani — Function.

Dixon-Maany function in size n (version I by default)

OptimizationProblems.dixmaanj — Function.

Dixon-Maany function in size n (version J)

OptimizationProblems.dixmaank — Function.

Dixon-Maany function in size n (version K)

OptimizationProblems.dixmaanl — Function.

Dixon-Maany function in size n (version L)

OptimizationProblems.dixmaanm — Function.

Dixon-Maany function in size n (version M by default)

OptimizationProblems.dixmaann — Function.

Dixon-Maany function in size n (version N)

OptimizationProblems.dixmaano — Function.

Dixon-Maany function in size n (version O)

OptimizationProblems.dixmaanp — Function.

Dixon-Maany function in size n (version P)

OptimizationProblems.dixon3dq — Function.

Dixon's tridiagonal quadratic.

OptimizationProblems.dqdrtic — Function.

Diagonal quadratic problem

OptimizationProblems.dqrtic — Function.

Diagonal quartic model in size n

OptimizationProblems.edensch — Function.

Extended Dennis-Schnabel model in size n

OptimizationProblems.eg2 — Function.

model in size n

OptimizationProblems.engval1 — Function.

The Engval1 model in size n

OptimizationProblems.errinros_mod

OptimizationProblems.eval

OptimizationProblems.extrosnb

OptimizationProblems.files

OptimizationProblems.fletcbv2

OptimizationProblems.fletcbv3_mod

OptimizationProblems.fletchcr

OptimizationProblems.fminsrf2

OptimizationProblems.freuroth

OptimizationProblems.genhumps

OptimizationProblems.genrose — Function.

Generalized Rosenbrock model in size n

OptimizationProblems.genrose_nash — Function.

Nash's variant of genrose() in size n

OptimizationProblems.hs1 — Function.

HS1 model

OptimizationProblems.hs10 — Function.

HS10 model

OptimizationProblems.hs100 — Function.

HS100 model

OptimizationProblems.hs101 — Function.

HS101 model

OptimizationProblems.hs102 — Function.

HS102 model

OptimizationProblems.hs103 — Function.

HS103 model

OptimizationProblems.hs104 — Function.

HS104 model

OptimizationProblems.hs105 — Function.

HS105 model

OptimizationProblems.hs106 — Function.

HS106 model

OptimizationProblems.hs107 — Function.

HS107 model

OptimizationProblems.hs108 — Function.

HS108 model

OptimizationProblems.hs109 — Function.

HS109 model

OptimizationProblems.hs11 — Function.

HS11 model

OptimizationProblems.hs110 — Function.

HS110 model

OptimizationProblems.hs111 — Function.

HS111 model

OptimizationProblems.hs112 — Function.

HS112 model

OptimizationProblems.hs113 — Function.

HS113 model

OptimizationProblems.hs114 — Function.

HS114 model

OptimizationProblems.hs116 — Function.

HS116 model

OptimizationProblems.hs117 — Function.

HS117 model

OptimizationProblems.hs118 — Function.

HS118 model

OptimizationProblems.hs119 — Function.

HS119 model

OptimizationProblems.hs12 — Function.

HS12 model

OptimizationProblems.hs13 — Function.

HS13 model

OptimizationProblems.hs14 — Function.

HS14 model

OptimizationProblems.hs15 — Function.

HS15 model

OptimizationProblems.hs16 — Function.

HS16 model

OptimizationProblems.hs17 — Function.

HS17 model

OptimizationProblems.hs18 — Function.

HS18 model

OptimizationProblems.hs19 — Function.

HS19 model

OptimizationProblems.hs2 — Function.

HS2 model

OptimizationProblems.hs20 — Function.

HS20 model

OptimizationProblems.hs21 — Function.

HS21 model

OptimizationProblems.hs22 — Function.

HS22 model

OptimizationProblems.hs23 — Function.

HS23 model

OptimizationProblems.hs24 — Function.

HS24 model

OptimizationProblems.hs25 — Function.

HS25 model

OptimizationProblems.hs26 — Function.

HS26 Model

OptimizationProblems.hs27 — Function.

HS27 Model

OptimizationProblems.hs28 — Function.

HS28 model

OptimizationProblems.hs29 — Function.

HS29 model

OptimizationProblems.hs3 — Function.

HS3 model

OptimizationProblems.hs30 — Function.

HS30 model

OptimizationProblems.hs31 — Function.

HS31 model

OptimizationProblems.hs32 — Function.

HS32 model

OptimizationProblems.hs33 — Function.

HS33 model

OptimizationProblems.hs34 — Function.

HS34 model

OptimizationProblems.hs35 — Function.

HS35 model

OptimizationProblems.hs36 — Function.

HS36 model

OptimizationProblems.hs37 — Function.

HS37 model

OptimizationProblems.hs38 — Function.

HS38 model

OptimizationProblems.hs39 — Function.

HS39 model

OptimizationProblems.hs4 — Function.

HS4 model

OptimizationProblems.hs40 — Function.

HS40 model

OptimizationProblems.hs41 — Function.

HS41 model

OptimizationProblems.hs42 — Function.

HS42 model

OptimizationProblems.hs43 — Function.

HS43 model

OptimizationProblems.hs44 — Function.

HS44 model

OptimizationProblems.hs45 — Function.

HS45 model

OptimizationProblems.hs46 — Function.

HS46 model

OptimizationProblems.hs47 — Function.

HS47 model

OptimizationProblems.hs48 — Function.

HS48 model

OptimizationProblems.hs49 — Function.

HS49 model

OptimizationProblems.hs5 — Function.

HS5 model

OptimizationProblems.hs50 — Function.

HS50 model

OptimizationProblems.hs51 — Function.

HS51 model

OptimizationProblems.hs52 — Function.

HS52 model

OptimizationProblems.hs53 — Function.

HS53 model

OptimizationProblems.hs54 — Function.

HS54 model

OptimizationProblems.hs55 — Function.

HS55 model

OptimizationProblems.hs56 — Function.

HS56 model

OptimizationProblems.hs57 — Function.

HS57 model

OptimizationProblems.hs59 — Function.

HS59 model

OptimizationProblems.hs6 — Function.

HS6 model

OptimizationProblems.hs60 — Function.

HS60 model

OptimizationProblems.hs61 — Function.

HS61 model

OptimizationProblems.hs62 — Function.

HS62 model

OptimizationProblems.hs63 — Function.

HS63 model

OptimizationProblems.hs64 — Function.

HS64 model

OptimizationProblems.hs65 — Function.

HS65 model

OptimizationProblems.hs66 — Function.

HS66 model

OptimizationProblems.hs7 — Function.

HS7 model

OptimizationProblems.hs70 — Function.

HS70 model

OptimizationProblems.hs71 — Function.

HS71 model

OptimizationProblems.hs72 — Function.

HS72 model

OptimizationProblems.hs73 — Function.

HS73 model

OptimizationProblems.hs74 — Function.

HS74 model

OptimizationProblems.hs75 — Function.

HS75 model

OptimizationProblems.hs76 — Function.

HS76 model

OptimizationProblems.hs77 — Function.

HS77 model

OptimizationProblems.hs78 — Function.

HS78 model

OptimizationProblems.hs79 — Function.

HS79 model

OptimizationProblems.hs8 — Function.

HS8 model

OptimizationProblems.hs80 — Function.

HS80 model

OptimizationProblems.hs81 — Function.

HS81 model

OptimizationProblems.hs83 — Function.

HS83 model

OptimizationProblems.hs84 — Function.

HS84 model

OptimizationProblems.hs86 — Function.

HS86 model

OptimizationProblems.hs9 — Function.

HS9 Model

OptimizationProblems.hs93 — Function.

HS93 model

OptimizationProblems.hs95 — Function.

HS95 model

OptimizationProblems.hs96 — Function.

HS96 model

OptimizationProblems.hs97 — Function.

HS97 model

OptimizationProblems.hs98 — Function.

HS98 model

OptimizationProblems.include

OptimizationProblems.indef_mod

OptimizationProblems.liarwhd

OptimizationProblems.meyer3 — Function.

Meyer function

OptimizationProblems.morebv

OptimizationProblems.nasty — Function.

Nasty problem

OptimizationProblems.ncb20

OptimizationProblems.ncb20b

OptimizationProblems.noncvxu2

OptimizationProblems.noncvxun

OptimizationProblems.nondia

OptimizationProblems.nondquar

OptimizationProblems.palmer1c — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer1d — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer2c — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer3c — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer4c — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer5c — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer5d — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer6c — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer7c — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer8c — Function.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.path

OptimizationProblems.penalty2 — Function.

A penalty problem by Gill, Murray and Pitfield in size 'n'

OptimizationProblems.penalty3 — Function.

A penalty problem by Gill, Murray and Pitfield in size 'n'

OptimizationProblems.powellsg — Function.

The extended Powell singular problem in size 'n'

OptimizationProblems.power — Function.

The Power problem by Oren.

OptimizationProblems.quartc — Function.

A simple quartic function.

OptimizationProblems.rosenbrock — Function.

Classic Rosenbrock problem

OptimizationProblems.sbrybnd — Function.

Broyden banded system of nonlinear equations in size 'n'

OptimizationProblems.schmvett — Function.

Another function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.scosine — Function.

Another function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.sinquad — Function.

Another function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.sparsine — Function.

A sparse problem involving sine functions in size 'n'

OptimizationProblems.sparsqur — Function.

A sparse quartic problem in size 'n'

OptimizationProblems.srosenbr — Function.

The separable extension of Rosenbrock's function 'n'

OptimizationProblems.tointgss — Function.

Toint's Gaussian problem in size 'n'

OptimizationProblems.tquartic — Function.

A quartic function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.tridia — Function.

Shanno's TRIDIA quadratic tridiagonal problem.

OptimizationProblems.vardim — Function.

Variable dimension problem.

OptimizationProblems.woods — Function.

The Woods function.

The extended Woods problem n

OptimizationProblems.zangwil3 — Function.

Zangwil3 Model