Docstrings

SimradEK60.EK60Ping

SimradEK60.FLOAT_TYPE

SimradEK60.POWER_MULTIPLIER

SimradEK60.R — Function.

R(r, s, T)

TVG range correction for Ex60

R is the corrected range (m)

r the uncorrected range (m)

s is the TvgRangeCorrectionOffset value, See Echoview documentation, Simrad Time Varied Gain (TVG) range correction, http://bit.ly/2pqzS2D

T is the sample thickness (m)

R(ping::EK60Ping; soundvelocity = nothing, rangecorrectionoffset=2)

Returns the corrected range (depth) of samples in ping.

SimradEK60.SimradEK60

SimradEK60.Sv — Function.

Sv(Pr, λ, G, Ψ, c, α, Pt, τ, Sa, R)

where:

R = the corrected range (m) - see TVG Range Correction

Pr = received power (dB re 1 W) - see Simrad EK numbers to Power

Pt = transmitted power (W)

α = absorption coefficient (dB/m)

G0 = transducer peak gain (non-dimensional) calculated as 10^(G/10)

G is the Transducer gain (dB re 1)

λ = wavelength (m) = c/f

f = frequency (Hz)

c = sound speed (m/s)

τ = transmit pulse duration (s) - also known as the TransmittedPulseLength

ψ = Equivalent Two-way beam angle (Steradians) calculated as 10^(Ψ/10)

Ψ is the Two-way beam angle (dB re 1 Steradian)

Sa = Simrad correction factor (dB re 1m−1) determined during calibration. This represents the correction required to the Sv constant to harmonize the TS and NASC measurements.

Sv(ping::EK60Ping;
        frequency=nothing,
        gain=nothing,
        equivalentbeamangle=nothing,
        soundvelocity=nothing,
        absorptioncoefficient=nothing,
        transmitpower=nothing,
        pulselength=nothing,
        sacorrection=nothing,
        rangecorrectionoffset=2)

Returns a Vector of Sv, the (Mean) Volume backscattering strength (MVBS) in (dB re 1 m-1) for a given ping.

The function accepts a number of optional arguments which, if specified, override the ping's own settings. This facilitates external calibration.

Sv(pings::Vector{EK60Ping};
    frequency=nothing,
    gain=nothing,
    equivalentbeamangle=nothing,
    soundvelocity=nothing,
    absorptioncoefficient=nothing,
    transmitpower=nothing,
    pulselength=nothing,
    sacorrection=nothing,
    rangecorrectionoffset=2)

Returns an Array of Sv, the Volume backscattering strength in (dB re 1 m-1) for a set of contiguous pings.

The function accepts a number of optional arguments which, if specified, override the pings' own settings. This facilitates external calibration.

SimradEK60.TS — Function.

 TS(Pr, λ, G, α, Pt, R)

Target strength

Pr = received power (dB re 1 W) - see Simrad EK numbers to Power

G is the Transducer gain (dB re 1)

α = absorption coefficient (dB/m)

Pt = transmitted power (W)

R = the corrected range (m) - see TVG Range Correction

TS(ping::EK60Ping;
        frequency=nothing,
        gain=nothing,
        soundvelocity=nothing,
        absorptioncoefficient=nothing,
        transmitpower=nothing,
        rangecorrectionoffset=0)

Target strength

TS(pings::Vector{EK60Ping};
        frequency=nothing,
        gain=nothing,
        soundvelocity=nothing,
        absorptioncoefficient=nothing,
        transmitpower=nothing,
        rangecorrectionoffset=0)

Target strength

SimradEK60.alongshipangle — Function.

alongshipangle(ping::EK60Ping)

Returns the alongship phase angle difference vector for the given ping.

SimradEK60.alongshipanglematrix

SimradEK60.athwartshipangle — Function.

athwartshipangle(ping::EK60Ping)

Returns the athwartship phase angle difference vector for the given ping.

SimradEK60.athwartshipanglematrix

SimradEK60.eval

SimradEK60.filetime — Function.

filetime(ping::EK60Ping)

Returns the FILETIME timestamp for a ping

SimradEK60.include

SimradEK60.load

SimradEK60.myhcat

SimradEK60.octet2deg — Function.

octet2deg(x)

EK60 split beam angles are stored as signed octets with a resolution of 180/128. Convert from octet to degrees.

SimradEK60.pings — Function.

pings(datagrams::Channel{SimradRaw.Datagram})

Returns a Channel of Pings from a Channel of Datagrams.

pings(filename::AbstractString)

Returns a Channel over Pings from the RAW file designated by filename.

pings(filenames::Vector{AbstractString}})

Returns a Channel of Pings from the RAW files designated by filenames.

SimradEK60.power — Function.

power(ping::EK60Ping)

Returns power data for a ping in manufacturer units. For decibels, see powerdb.

SimradEK60.power2db — Function.

power2db(x)

EK60 power data is stored in a "compressed", manufacturer specific format. This function converts to decibels.

SimradEK60.powerdb — Function.

powerdb(x)

Returns power data in decibels. x is one or more Pings.