Docstrings

SemialgebraicSets.@set

SemialgebraicSets.APL

SemialgebraicSets.AbstractAlgebraicSet

SemialgebraicSets.AbstractAlgebraicSolver — Type.

AbstractAlgebraicSolver

Solver of algebraic equations.

SemialgebraicSets.AbstractBasicSemialgebraicSet

SemialgebraicSets.AbstractGröbnerBasisAlgorithm

SemialgebraicSets.AbstractMultiplicationMatricesAlgorithm — Type.

AbstractMultiplicationMatricesAlgorithm

Algorithm computing multiplication matrices from algebraic equations.

SemialgebraicSets.AbstractMultiplicationMatricesSolver — Type.

AbstractMultiplicationMatricesSolver

Solver of algebraic equations using multiplication matrices.

SemialgebraicSets.AbstractPolynomialIdeal

SemialgebraicSets.AbstractSemialgebraicSet

SemialgebraicSets.AlgebraicSet

SemialgebraicSets.BasicSemialgebraicSet

SemialgebraicSets.Buchberger

SemialgebraicSets.DefaultAlgebraicSetLibrary

SemialgebraicSets.EmptyPolynomialIdeal

SemialgebraicSets.FullSpace

SemialgebraicSets.GröbnerBasisMultiplicationMatricesAlgorithm

SemialgebraicSets.MP

SemialgebraicSets.PolynomialIdeal

SemialgebraicSets.ReorderedSchurMultiplicationMatricesSolver — Type.

Corless, R. M.; Gianni, P. M. & Trager, B. M. A reordered Schur factorization method for zero-dimensional polynomial systems with multiple roots Proceedings of the 1997 international symposium on Symbolic and algebraic computation, 1997, 133-140

SemialgebraicSets.SemialgebraicSets

SemialgebraicSets.SolverUsingMultiplicationMatrices

SemialgebraicSets._canonicalize_sense

SemialgebraicSets._clusterordschur

SemialgebraicSets._solvemultiplicationmatrices

SemialgebraicSets._trsen!

SemialgebraicSets.addequality!

SemialgebraicSets.addinequality!

SemialgebraicSets.algebraicset

SemialgebraicSets.algebraicsolver

SemialgebraicSets.appendconstraints!

SemialgebraicSets.basicsemialgebraicset

SemialgebraicSets.builddomain

SemialgebraicSets.clusterordschur

SemialgebraicSets.computeelements!

SemialgebraicSets.computegröbnerbasis!

SemialgebraicSets.conditionnumber

SemialgebraicSets.criterion

SemialgebraicSets.defaultalgebraicsetlibrary

SemialgebraicSets.defaultalgebraicsolver

SemialgebraicSets.defaultgröbnerbasisalgorithm

SemialgebraicSets.defaultmultiplicationmatricesalgorithm

SemialgebraicSets.defaultmultiplicationmatricessolver

SemialgebraicSets.dummyselection

SemialgebraicSets.elements

SemialgebraicSets.equalities

SemialgebraicSets.eval

SemialgebraicSets.ext

SemialgebraicSets.groebnerbasis

SemialgebraicSets.gröbnerbasis

SemialgebraicSets.gröbnerbasis!

SemialgebraicSets.ideal

SemialgebraicSets.include

SemialgebraicSets.inequalities

SemialgebraicSets.iszerodimensional

SemialgebraicSets.lcmlm

SemialgebraicSets.monomialbasis

SemialgebraicSets.multiplicationmatrices — Function.

multiplicationmatrices(V::AbstractAlgebraicSet, algo::AbstractMultiplicationMatricesAlgorithm)::Union{Nothing, Vector{<:AbstractMatrix}}

Computing multiplication matrices from the algebraic equations for which V is the set of solution using the algorithm algo. Returns a nullable which is null if V is not zero-dimensional and is the list of multiplication matrices otherwise.

SemialgebraicSets.multiplicationmatrix

SemialgebraicSets.nequalities

SemialgebraicSets.normalselection

SemialgebraicSets.presort!

SemialgebraicSets.projectivealgebraicset

SemialgebraicSets.reducebasis!

SemialgebraicSets.solvealgebraicequations — Function.

solvealgebraicequations(V::AbstractAlgebraicSet, algo::AbstractAlgebraicSolver)::Union{Nothing, Vector{<:Vector}}}

Solve the algebraic equations for which V is the set of solutions using the algorithm algo. Returns a nullable which is null if V is not zero-dimensional and is the list of solutions otherwise.

SemialgebraicSets.solvemultiplicationmatrices — Function.

solvemultiplicationmatrices(Ms::AbstractVector{<:AbstractMatrix{T}}, algo::AbstractMultiplicationMatricesSolver)::Vector{Vector{T}} where T

Solve the algebraic equations having multiplication matrices Ms using the algorithm algo. Returns the list of solutions.

SemialgebraicSets.spolynomial — Function.

spolynomial(p::AbstractPolynomialLike, q::AbstractPolynomialLike)

Computes the S-polynomial of p and q defined by

\[S(p, q) = \frac{m}{\mathrm{\mathsc{LT}}(p)} \cdot p - \frac{m}{\mathrm{\mathsc{LT}}(q)} \cdot q\]

where $m = \mathrm{lcm}(\mathrm{\mathsc{LM}}(p), \mathrm{\mathsc{LM}}(q))$.