Docstrings

RestrictProlong.DimsLike

RestrictProlong.RestrictProlong

RestrictProlong._prolong

RestrictProlong._prolong!

RestrictProlong._restrict

RestrictProlong._restrict!

RestrictProlong._size

RestrictProlong.checkdims

RestrictProlong.eval

RestrictProlong.include

RestrictProlong.prolong — Function.

Ap = prolong(A, [sz::Tuple])

Perform two-fold expansion in size. If sz is omitted, all dimensions are expanded. If sz is specified, it can be one of the following:

sz[d] = size(A, d): no expansion is performed

sz[d] = 2*size(A, d) - 1: expansion to an odd-size, adding (by interpolation) points on the half-grid between all points along dimension d

sz[d] = 2*size(A, d): expansion to an even-size, estimating the points on the 1/4 and 3/4 grid (including one point beyond either edge of the current span)

Any other choice results in an error. The default choice is to perform odd-expansion. prolong is normalized so as to approximately preserve the sum of A.

Thought of as an operator, prolong is equal to the transpose of restrict.

prolong(A, sz::Integer, dim::Integer)

Perform expansion to size sz along dimension dim. sz must be one of the valid choices for prolongation size.

RestrictProlong.restrict — Function.

Ar = restrict(A[, dims])

Perform two-fold reduction in size along the dimensions listed in dims, or all coordinates if dims is omitted. It anti-aliases A, so is more accurate than a naive summation over 2×2×... blocks. restrict is normalized so as to approximately preserve the mean value of A.

Thought of as an operator, restrict is equal to the transpose of prolong.

RestrictProlong.restrict!

RestrictProlong.restrict_size