Docstrings

AbstractInstances.AbstractInstances

AbstractInstances._concretesymbol

AbstractInstances._new_concrete

AbstractInstances._new_singleton

AbstractInstances._singletonsymbol

AbstractInstances.concretetype — Function.

concretetype(T::Type)

Return a concrete mutable type with no fields which is a subtype of T. The mutability of the returned type ensures that instances of the type are unique.

AbstractInstances.eval

AbstractInstances.include

AbstractInstances.oftype — Function.

oftype(T::Type)

Create a unique instance of type, concretetype(T).

AbstractInstances.singleton — Function.

singleton(T::Type)

Create an instance of type, singletontype(T).

AbstractInstances.singletontype — Function.

singletontype(T::Type)

Return a concrete immutable type with no fields which is a subtype of T. The immutability of the returned type ensures that instances of the type are singletons.