Docstrings

CutPruners.AbstractCutPruner — Type.

A cut pruner maintains a matrix A and a vector b such that represents size(A, 1) (== length(b)) cuts. Let a_i be -A[i,:] if lazy_minus and A[i,:] otherwise and β_i be b[i], the meaning of the cut depends on the sense. If sense is

:Min, then the cut pruner represents the concave polyhedral function min ⟨a_i, x⟩ + β_i;

:Max, then the cut pruner represents the convex polyhedral function max ⟨a_i, x⟩ + β_i;

:≤, then the cut pruner represents the polyhedra defined by the intersection of the half-space ⟨a_i, x⟩ ≤ β_i;

:≥, then the cut pruner represents the polyhedra defined by the intersection of the half-space ⟨a_i, x⟩ ≥ β_i.

Internally, instead of sense, the booleans isfun and islb are stored. The mapping between sense and these two booleans is given by the following table

sense	isfun	islb

Min	true	false

Max	true	true

≤	false	false

≥	false	true

CutPruners.AbstractCutPruningAlgo

CutPruners.AvgCutPruner

CutPruners.AvgCutPruningAlgo — Type.

AvgCutPruningAlgo <: AbstractCutPruningAlgo

Removes the cuts with lower trust where the trust is: nused / nwith + bonus where the cut has been used nused times amoung nwith optimization done with it. We say that the cut was used if its dual value is nonzero. It has a bonus equal to mycutbonus if the cut was generated using a trial given by the problem using this cut. If nwidth is zero, nused/nwith is replaced by newcuttrust.

CutPruners.CutPruner

CutPruners.CutPruners

CutPruners.DeMatosCutPruner

CutPruners.DeMatosPruningAlgo — Type.

DeMatosPruningAlgo <: AbstractCutPruningAlgo

Removes the cuts with lower trust where the trust is the number of points x associated to the cuts. The more points are associated, the higher is the trust.

We refer to [1] for further details.

[1] De Matos, Vitor L., Andy B. Philpott, and Erlon C. Finardi. "Improving the performance of stochastic dual dynamic programming." Journal of Computational and Applied Mathematics 290 (2015): 196-208.

CutPruners.DecayCutPruner

CutPruners.DecayCutPruningAlgo — Type.

DecayCutPruningAlgo <: AbstractCutPruningAlgo

Removes the cuts with lower trust where the trust is initially newcuttrust + bonus and is updated using trust -> λ * trust + used after each optimization done with it. The value used is 1 if the cut was used and 0 otherwise. It has a bonus equal to mycutbonus if the cut was generated using a trial given by the problem using this cut. We say that the cut was used if its dual value is nonzero.

CutPruners._appendcuts!

CutPruners._indmin

CutPruners._keeponlycuts!

CutPruners._replacecuts!

CutPruners.addcuts!

CutPruners.addposition! — Function.

addposition!(man::DeMatosCutPruner, position::Matrix)

Update territories with cuts previously computed during backward pass.

Arguments

man::DeMatosCutPruner Pruner to update.

position::Array{T, 2} New visited positions, corresponding to a collection of points.

CutPruners.addstate! — Function.

addstate!(man::DeMatosCutPruner, x::Vector)

Add a new state to test and accordingly update territories of each cut.

CutPruners.addusage! — Function.

Update cuts relevantness after a solver's call returning dual vector σρ.

CutPruners.appendcuts! — Function.

Append cuts (A, b, mycut) in CutPruner man.

Push new cut in CutPruner man.

CutPruners.checkredundancy — Function.

checkredundancy{T}(A::AbstractMatrix{T}, b::AbstractVector{T},
                   Anew::AbstractMatrix{T}, bnew::AbstractVector{T},
                   isfun::Bool, islb::Bool, tol::Float64, ident::Bool=false)

Check redundant cuts between the Polyhedra (A,b) and (Anew, bnew). Return index of redundant cuts in Anew.

Arguments

isfun::Bool States if the Polyhedra defines a function

islb::Bool States if the Polyhedra is a lower-bound or an upper-bound

tol::Float64 Tolerance of redundancy check

ident::Bool States whether (A,b)==(Anew,bnew) if we want to remove redundant lines in a single Polyhedra

CutPruners.choosecutstoremove — Function.

Remove num cuts in CutPruner man.

CutPruners.cutvalue — Function.

cutvalue{T}(man::DeMatosCutPruner, indc::Int, x::Vector{T})

Get value of cut with index indc at point x.

Arguments

man::DeMatosCutPruner Approximation of the value function as linear cuts

indc::Int Index of cut to consider

x::Vector{T} Coordinates of state

Return

cost::T Value of cut indc at point x. If man is a polyhedral function, then it is the value of the cut at x, otherwise, it is the distance between x and the cut. As a rule of thumb, the higher the cutvalue is, the less it is redundant.

CutPruners.eval

CutPruners.exactpruning! — Function.

exactpruning!(man::AbstractCutPruner, solver::MathProgBase.AbstractMathProgSolver;
              ub=Inf, lb=-Inf, epsilon=1e-5)

Remove dominated cuts in CutPruner man.

We use a LP solver to determine whether a cut is dominated or not.

Arguments

man::AbstractCutPruner Cut pruner where to remove cuts

solver Solver used to solve LP

ub::Union{Float64, Vector{Float64}} State x upper bound

lb::Union{Float64, Vector{Float64}} State x lower bound

epsilon::Float64 Pruning's tolerance

CutPruners.flength

CutPruners.getdominated — Function.

Return dominated cuts.

CutPruners.getnreplaced

CutPruners.getsense

CutPruners.gettrust — Function.

Get current trust of CutPruner man.

CutPruners.gettrustof

CutPruners.gettype

CutPruners.giveterritory!

CutPruners.hasterritories

CutPruners.hastrust

CutPruners.include

CutPruners.initialtrust

CutPruners.initialtrusts — Function.

Get a Vector of Float64 specifying the initial trusts of mycut.

CutPruners.isbetter — Function.

Test if cut i is better than newcuttrust.

CutPruners.isdominated — Function.

State whether a cut is dominated with a tolerance epsilon.

CutPruners.isfun

CutPruners.isinside — Function.

Check if λ is a line of matrix A. λ might not have the same eltype as A and b as it might have been scaled by normalizecut.

CutPruners.islb

CutPruners.keeponlycuts! — Function.

Keep only cuts whose indexes are in Vector K in CutPruner man. If K is not sorted, the cuts will change their order accordingly.

CutPruners.ncuts — Function.

Return number of cuts in CutPruner man.

CutPruners.newids

CutPruners.normalizedcut

CutPruners.optimalcut — Function.

optimalcut{T}(man::DeMatosCutPruner, xf::Vector{T})

Find active cut at point xf.

Arguments

man::DeMatosCutPruner: CutPruner

xf::Vector{T}:

Return

bestcost::T Value of supporting cut at point xf

bestcut::Int Index of supporting cut at point xf

CutPruners.removecuts! — Function.

Remove cuts whose indexes are in Vector K in CutPruner man.

CutPruners.replacecuts! — Function.

Replace cuts at indexes in K by cuts in (A, b, mycut) in CutPruner man.

CutPruners.updateterritory! — Function.

updateterritory!(man::DeMatosCutPruner, indcut::Int)

Update territories (i.e. the set of tested states where a given cut is active) considering new cut given by index indcut.

Arguments

man::DeMatosCutPruner:

indcut::Int: new cut index

CutPruners.updatetrust!