Docstrings

SummationByPartsOperators.AbstractAnalyticalProblem

SummationByPartsOperators.AbstractCoefficientCache

SummationByPartsOperators.AbstractDerivativeCoefficients

SummationByPartsOperators.AbstractDerivativeOperator

SummationByPartsOperators.AbstractFilter

SummationByPartsOperators.AbstractFilterFunction

SummationByPartsOperators.AbstractFourierViscosity

SummationByPartsOperators.AbstractMassMatrix

SummationByPartsOperators.AbstractPeriodicDerivativeOperator

SummationByPartsOperators.AbstractSemidiscretisation

SummationByPartsOperators.AbstractVariableCoefficientDerivativeOperator

SummationByPartsOperators.AffineDiffEqOperator

SummationByPartsOperators.AnalyticalProblem

SummationByPartsOperators.BVProblem

SummationByPartsOperators.BeljaddLeFlochMishraParés2017 — Type.

BeljaddLeFlochMishraParés2017

Coefficients of the periodic operators given in Beljadid, LeFloch, Mishra, Parés (2017) Schemes with Well-Controlled Dissipation. Hyperbolic Systems in Nonconservative Form. Communications in Computational Physics 21.4, pp. 913-946.

SummationByPartsOperators.BurgersNonperiodicSemidiscretisation — Type.

BurgersNonperiodicSemidiscretisation

A semidiscretisation of Burgers' equation $\partial_t u(t,x) + \partial_x \frac{u(t,x)^2}{2} = 0$ with boundary conditions left_bc(t), right_bc(t).

SummationByPartsOperators.BurgersPeriodicSemidiscretisation — Type.

BurgersPeriodicSemidiscretisation

A semidiscretisation of Burgers' equation $\partial_t u(t,x) + \partial_x \frac{u(t,x)^2}{2} = 0$ with periodic boundary conditions.

SummationByPartsOperators.CallbackSet

SummationByPartsOperators.ClenshawCurtis

SummationByPartsOperators.ClosedNewtonCotes

SummationByPartsOperators.ConstantFilter — Type.

ConstantFilter

Represents the action of a modal filter on values in a nodal basis with fixed strength.

SummationByPartsOperators.ContinuousCallback

SummationByPartsOperators.CubicNonperiodicSemidiscretisation — Type.

CubicNonperiodicSemidiscretisation

A semidiscretisation of the cubic conservation law $\partial_t u(t,x) + \partial_x u(t,x)^3 = 0$ with nonperiodic boundary conditions.

SummationByPartsOperators.CubicPeriodicSemidiscretisation — Type.

CubicPeriodicSemidiscretisation{T,Derivative,Dissipation}

A semidiscretisation of the cubic conservation law $\partial_t u(t,x) + \partial_x u(t,x)^3 = 0$ with periodic boundary conditions.

SummationByPartsOperators.DAEFunction

SummationByPartsOperators.DAEProblem

SummationByPartsOperators.DAESolution

SummationByPartsOperators.DDEFunction

SummationByPartsOperators.DDEProblem

SummationByPartsOperators.DEDataArray

SummationByPartsOperators.DEDataMatrix

SummationByPartsOperators.DEDataVector

SummationByPartsOperators.DEFAULT_LINSOLVE

SummationByPartsOperators.DESolution

SummationByPartsOperators.DerivativeCoefficientRow — Type.

DerivativeCoefficientRow{T,Start,Length}

A struct representing a row in the boundary block of an SBP derivative operator with scalar type T.

SummationByPartsOperators.DerivativeCoefficients — Type.

DerivativeCoefficients

The coefficients of a derivative operator on a nonperiodic grid.

SummationByPartsOperators.DerivativeOperator — Type.

DerivativeOperator

A derivative operator on a nonperiodic finite difference grid.

SummationByPartsOperators.DiffEqBase

SummationByPartsOperators.DiscreteCallback

SummationByPartsOperators.DiscreteFunction

SummationByPartsOperators.DiscreteProblem

SummationByPartsOperators.DissipationOperator — Type.

DissipationOperator

A dissipation operator on a nonperiodic finite difference grid.

SummationByPartsOperators.DynamicalODEFunction

SummationByPartsOperators.DynamicalODEProblem

SummationByPartsOperators.ExponentialFilter — Type.

ExponentialFilter

Represents the exponential filter function σ(η) = exp(-α*η^p).

SummationByPartsOperators.FactorisationWrapper — Type.

FactorisationWrapper

A small wrapper around a a factorisation fact, allowing to represent multiplication with the inverse of fact.

SummationByPartsOperators.FilterCallback

SummationByPartsOperators.Fornberg1998 — Type.

Fornberg1998

Coefficients of the periodic operators given in Fornberg (1998) Calculation of Weights in Finite Difference Formulas. SIAM Rev. 40.3, pp. 685-691.

SummationByPartsOperators.FourierConstantViscosity — Type.

FourierConstantViscosity

Fourier viscosity operator with constant coefficients for the periodic 1st derivative Fourier operator.

SummationByPartsOperators.FourierDerivativeOperator — Type.

FourierDerivativeOperator{T<:Real, Grid, RFFT, BRFFT}

A derivative operator on a periodic grid with scalar type T computing the first derivative using a spectral Fourier expansion via real discrete Fourier transforms.

SummationByPartsOperators.FourierSpectralViscosityCoefficients

SummationByPartsOperators.FourierSuperSpectralViscosityCoefficients — Type.

Tadmor1993

Coefficients of the Fourier super spectral viscosity given in Tadmor (1993) Super Viscosity and Spectral Approximations of Nonlinear Conservation Laws. Numerical Methods for Fluid Dynamics IV, pp. 69-82.

SummationByPartsOperators.GaussJacobi

SummationByPartsOperators.GaussLegendre

SummationByPartsOperators.Holoborodko2008 — Type.

Holoborodko2008

Coefficients of the periodic operators given in Holoborodko (2008) Smooth Noise Robust Differentiators. http://www.holoborodko.com/pavel/numerical-methods/numerical-derivative/smooth-low-noise-differentiators/

SummationByPartsOperators.LegendreDerivativeOperator — Type.

LegendreDerivativeOperator{T<:Real, Grid}

A derivative operator on a nonperiodic Lobatto-Legendre grid with scalar type T computing the first derivative using a Legendre expansion.

SummationByPartsOperators.LinSolveFactorize

SummationByPartsOperators.LobattoLegendre

SummationByPartsOperators.MadayTadmor1989 — Type.

MadayTadmor1989

Coefficients of the Fourier spectral viscosity given in Maday, Tadmor (1989) Analysis of the Spectral Vanishing Viscosity Method for Periodic Conservation Laws. SIAM Journal on Numerical Analysis 26.4, pp. 854-870.

SummationByPartsOperators.Mattsson2012 — Type.

Mattsson2012

Coefficients of the SBP operators given in Mattsson (2012) Summation by Parts Operators for Finite Difference Approximations of Second-Derivatives with Variable Coefficients. Journal of Scientific Computing 51, pp. 650-682.

SummationByPartsOperators.Mattsson2012Cache2

SummationByPartsOperators.Mattsson2012Cache4

SummationByPartsOperators.Mattsson2012Cache6

SummationByPartsOperators.Mattsson2014 — Type.

Mattsson2014

Coefficients of the SBP operators given in Mattsson (2014) Diagonal-norm summation by parts operators for fiite difference approximations of third and fourth derivatives. Journal of Computational Physics 274, pp. 432-454.

SummationByPartsOperators.MattssonAlmquistCarpenter2014Extended — Type.

MattssonAlmquistCarpenter2014Extended

Coefficients of the extended SBP operators given in Mattsson, Almquist, Carpenter (2014) Optimal diagonal-norm SBP operators. Journal of Computational Physics 264, pp. 91-111.

SummationByPartsOperators.MattssonAlmquistCarpenter2014Optimal — Type.

MattssonAlmquistCarpenter2014Optimal

Coefficients of the optimal SBP operators with nonuniform grid given in Mattsson, Almquist, Carpenter (2014) Optimal diagonal-norm SBP operators. Journal of Computational Physics 264, pp. 91-111.

SummationByPartsOperators.MattssonAlmquistCarpenter2014OptimalGrid

SummationByPartsOperators.MattssonNordström2004 — Type.

MattssonNordström2004

Coefficients of the SBP operators given in Mattsson, Nordström (2004) Summation by parts operators for finite difference approximations of second derivatives. Journal of Computational Physics 199, pp.503-540.

SummationByPartsOperators.MattssonSvärdNordström2004 — Type.

MattssonSvärdNordström2004

Coefficients of the SBP operators given in Mattsson, Svärd, Nordström (2004) Stable and Accurate Artificial Dissipation. Journal of Scientific Computing 21.1, pp. 57-79.

SummationByPartsOperators.MattssonSvärdNordström2004Cache2

SummationByPartsOperators.MattssonSvärdNordström2004Cache4

SummationByPartsOperators.MattssonSvärdNordström2004Cache6

SummationByPartsOperators.MattssonSvärdNordström2004Cache8

SummationByPartsOperators.MattssonSvärdShoeybi2008 — Type.

MattssonSvärdShoeybi2008

Coefficients of the SBP operators given in Mattsson, Svärd, Shoeybi (2008) Stable and accurate schemes for the compressible Navier-Stokes equations. Journal of Computational Physics 227, pp. 2293-2316.

SummationByPartsOperators.MonteCarloProblem

SummationByPartsOperators.MonteCarloSolution

SummationByPartsOperators.MonteCarloSummary

SummationByPartsOperators.MonteCarloTestSolution

SummationByPartsOperators.NodalBasis

SummationByPartsOperators.NoiseProblem

SummationByPartsOperators.ODEFunction

SummationByPartsOperators.ODEProblem

SummationByPartsOperators.ODESolution

SummationByPartsOperators.PeriodicDerivativeCoefficients — Type.

PeriodicDerivativeCoefficients{T,LowerOffset,UpperOffset,Parallel}

The coefficients of a derivative operator on a periodic grid.

SummationByPartsOperators.PeriodicDerivativeOperator — Type.

PeriodicDerivativeOperator

A derivative operator on a uniform periodic grid.

SummationByPartsOperators.PeriodicDissipationOperator — Type.

PeriodicDissipationOperator

A dissipation operator on a periodic finite difference grid.

SummationByPartsOperators.PolynomialBases

SummationByPartsOperators.RODEFunction

SummationByPartsOperators.RODEProblem

SummationByPartsOperators.RODESolution

SummationByPartsOperators.SDEFunction

SummationByPartsOperators.SDEProblem

SummationByPartsOperators.ScalarIntegralQuantities

SummationByPartsOperators.SecondOrderODEProblem

SummationByPartsOperators.SourceOfCoefficients — Type.

SourceOfCoefficients

All sources of coefficients (articles) are subtypes of this abstract type.

SummationByPartsOperators.SplitFunction

SummationByPartsOperators.SplitODEProblem

SummationByPartsOperators.SplitSDEFunction

SummationByPartsOperators.SplitSDEProblem

SummationByPartsOperators.SteadyStateProblem

SummationByPartsOperators.SteadyStateSolution

SummationByPartsOperators.SumOfDerivativeOperators — Type.

SumOfDerivativeOperators

Sum several derivative operators lazily.

SummationByPartsOperators.SummationByPartsOperators

SummationByPartsOperators.Tadmor1989 — Type.

Tadmor1989

Coefficients of the Fourier spectral viscosity given in Tadmor (1989) Convergence of Spectral Methods for Nonlinear Conservation Laws. SIAM Journal on Numerical Analysis 26.1, pp. 30-44.

SummationByPartsOperators.Tadmor1993 — Type.

Tadmor1993

Coefficients of the Fourier super spectral viscosity given in Tadmor (1993) Super Viscosity and Spectral Approximations of Nonlinear Conservation Laws. Numerical Methods for Fluid Dynamics IV, pp. 69-82.

SummationByPartsOperators.TadmorWaagan2012Convergent — Type.

TadmorWaagan2012Convergent

Coefficients of the Fourier spectral viscosity given in Tadmor, Waagan (2012) Adaptive Spectral Viscosity for Hyperbolic Conservation Laws. SIAM Journal on Scientific Computing 34.2, pp. A993-A1009. See also Schochet (1990) The Rate of Convergence of Spectral-Viscosity Methods for Periodic Scalar Conservation Laws. SIAM Journal on Numerical Analysis 27.5, pp. 1142-1159.

SummationByPartsOperators.TadmorWaagan2012Standard — Type.

TadmorWaagan2012Standard

Coefficients of the Fourier spectral viscosity given in Tadmor, Waagan (2012) Adaptive Spectral Viscosity for Hyperbolic Conservation Laws. SIAM Journal on Scientific Computing 34.2, pp. A993-A1009.

SummationByPartsOperators.TimeChoiceIterator

SummationByPartsOperators.TwoPointBVProblem

SummationByPartsOperators.VarCoefDerivativeCoefficients — Type.

VarCoefDerivativeCoefficients

The coefficients of a variable coefficient derivative operator on a nonperiodic grid.

SummationByPartsOperators.VarCoefDerivativeOperator — Type.

VarCoefDerivativeOperator

A dissipation operator on a nonperiodic finite difference grid.

SummationByPartsOperators.VariableLinearAdvectionNonperiodicSemidiscretisation — Type.

VariableLinearAdvectionNonperiodicSemidiscretisation

A semidiscretisation of the linear advection equation $\partial_t u(t,x) + \partial_x ( a(x) u(t,x) ) = 0$ with boundary conditions left_bc(t), right_bc(t).

SummationByPartsOperators.WaveEquationNonperiodicSemidiscretisation — Type.

WaveEquationNonperiodicSemidiscretisation

A semidiscretisation of the linear wave equation $\partial_t^2 u(t,x) = \partial_x^2 u(t,x)$.

SummationByPartsOperators.__init__

SummationByPartsOperators.accuracy_order

SummationByPartsOperators.add_saveat!

SummationByPartsOperators.add_transpose_derivative_left! — Function.

addtransposederivativeleft!(u, D::DerivativeOperator, derorder::Val{N})

Compute the transposed N-th derivative to the function given by the coefficients u at the left boundary of the grid.

SummationByPartsOperators.add_transpose_derivative_right! — Function.

addtransposederivativeright!(u, D::DerivativeOperator, derorder::Val{N})

Compute the transposed N-th derivative to the function given by the coefficients u at the right boundary of the grid.

SummationByPartsOperators.add_tstop!

SummationByPartsOperators.addat!

SummationByPartsOperators.addat_non_user_cache!

SummationByPartsOperators.addsteps!

SummationByPartsOperators.auto_dt_reset!

SummationByPartsOperators.change_basis

SummationByPartsOperators.change_basis!

SummationByPartsOperators.change_t_via_interpolation!

SummationByPartsOperators.check_error

SummationByPartsOperators.check_keywords

PolynomialBases.compute_coefficients — Function.

compute_coefficients(u, D::AbstractDerivativeOperator)

Compute the nodal values of the function u at the grid associated to the derivative operator D.

compute_coefficients(u, D::AbstractPeriodicDerivativeOperator)

Compute the nodal values of the function u at the grid associated to the derivative operator D.

PolynomialBases.compute_coefficients! — Function.

compute_coefficients!(uval::AbstractVector, u, D::AbstractDerivativeOperator)

Compute the nodal values of the function u at the grid associated to the derivative operator D and stores the result in uval.

compute_coefficients!(uval::AbstractVector, u, D::AbstractPeriodicDerivativeOperator)

Compute the nodal values of the function u at the grid associated to the derivative operator D and stores the result in uval.

SummationByPartsOperators.construct_grid

SummationByPartsOperators.convolve_boundary_coefficients!

SummationByPartsOperators.convolve_boundary_coefficients!_unrolled_expansion_

SummationByPartsOperators.convolve_interior_coefficients!

SummationByPartsOperators.convolve_interior_coefficients_loopbody

SummationByPartsOperators.convolve_left_row

SummationByPartsOperators.convolve_periodic_boundary_coefficients!

SummationByPartsOperators.convolve_right_row

SummationByPartsOperators.deleteat!

SummationByPartsOperators.deleteat_non_user_cache!

SummationByPartsOperators.derivative_at

SummationByPartsOperators.derivative_at!

SummationByPartsOperators.derivative_left — Function.

derivative_left(D::DerivativeOperator, u, der_order::Val{N})

Compute the N-th derivative of the function given by the coefficients u at the left boundary of the grid.

SummationByPartsOperators.derivative_matrix

SummationByPartsOperators.derivative_matrix!

SummationByPartsOperators.derivative_operator — Function.

derivative_operator(source_of_coefficients, derivative_order, accuracy_order,
                    xmin, xmax, N, parallel=Val{:serial}())

Create a DerivativeOperator approximating the derivative_order-th derivative on a grid between xmin and xmax with N grid points up to order of accuracy accuracy_order. with coefficients given by source_of_coefficients. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

SummationByPartsOperators.derivative_order

SummationByPartsOperators.derivative_right — Function.

derivative_right(D::DerivativeOperator, u, der_order::Val{N})

Compute the N-th derivative of the function given by the coefficients u at the right boundary of the grid.

SummationByPartsOperators.dissipation_coefficients

SummationByPartsOperators.dissipation_operator — Function.

dissipation_operator(D::PeriodicDerivativeOperator;
                     strength=one(eltype(D)),
                     order=accuracy_order(D),
                     parallel=D.coefficients.parallel)

Create a negative semidefinite DissipationOperator using undivided differences approximating a order-th derivative with strength strength adapted to the derivative operator D. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

dissipation_operator(source_of_coefficients, order, xmin, xmax, N, 
                     left_weights, right_weights, parallel=Val{:serial}())

Create a negative semidefinite DissipationOperator using undivided differences approximating a weighted order-th derivative on a grid between xmin and xmax with N grid points up to order of accuracy 2 with coefficients given by source_of_coefficients. The norm matrix is given by left_weights and right_weights. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

dissipation_operator(source_of_coefficients, D::DerivativeOperator{T};
                     strength=one(T),
                     order::Int=accuracy_order(D), 
                     parallel=D.coefficients.parallel)

Create a negative semidefinite DissipationOperator using undivided differences approximating a weighted order-th derivative adapted to the derivative operator D with coefficients given in source_of_coefficients. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

dissipation_operator(D::DerivativeOperator; kwargs...)

Create a negative semidefinite DissipationOperator using undivided differences approximating a weighted order-th derivative adapted to the derivative operator D. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

SummationByPartsOperators.du_cache

SummationByPartsOperators.eval

PolynomialBases.evaluate_coefficients — Function.

evaluate_coefficients(u, D::AbstractDerivativeOperator)

Evaluates the nodal coefficients u at a grid associated to the derivative operator D. Returns xplot, uplot, where xplot contains the nodes and uplot the corresponding values of u.

evaluate_coefficients(u, D::AbstractPeriodicDerivativeOperator)

Evaluates the nodal coefficients u at a grid including both endpoints associated to the derivative periodic operator D. Returns xplot, uplot, where xplot contains the equally spaced nodes and uplot the corresponding values of u.

PolynomialBases.evaluate_coefficients! — Function.

evaluate_coefficients!(xplot, uplot, u, D::AbstractDerivativeOperator)

Evaluates the nodal coefficients u at a grid associated to the derivative operator D and stores the result in xplot, uplot. Returns xplot, uplot, where xplot contains the nodes and uplot the corresponding values of u.

evaluate_coefficients!(xplot, uplot, u, D::AbstractPeriodicDerivativeOperator)

Evaluates the nodal coefficients u at a grid including both endpoints associated to the derivative periodic operator D and stores the result in xplot, uplot. Returns xplot, uplot, where xplot contains the equally spaced nodes and uplot the corresponding values of u.

SummationByPartsOperators.first_derivative_coefficients

SummationByPartsOperators.fornberg — Function.

fornberg(x::Vector{T}, m::Int) where {T}

Calculate the weights of a finite difference approximation of the mth derivative with maximal order of accuracy at 0 using the nodes x, see Fornberg (1998) Calculation of Weights in Finite Difference Formulas SIAM Rev. 40.3, pp. 685-691.

SummationByPartsOperators.fourier_derivative_matrix — Function.

fourier_derivative_matrix(N, xmin::Real=0.0, xmax::Real=2π)

Compute the Fourier derivative matrix with respect to the corresponding nodal basis using N nodes, see Kopriva (2009) Implementing Spectral Methods for PDEs, Algorithm 18.

SummationByPartsOperators.fourier_derivative_operator

SummationByPartsOperators.fourth_derivative_coefficients

SummationByPartsOperators.full_cache

SummationByPartsOperators.gegenbauer

SummationByPartsOperators.get_dt

SummationByPartsOperators.get_du

SummationByPartsOperators.get_du!

SummationByPartsOperators.get_parameters

SummationByPartsOperators.get_proposed_dt

SummationByPartsOperators.get_tmp_cache

SummationByPartsOperators.godunov_flux_burgers

SummationByPartsOperators.godunov_flux_cubic

SummationByPartsOperators.godunov_flux_variablelinearadvection

SummationByPartsOperators.grid

SummationByPartsOperators.hahn

SummationByPartsOperators.has_exp

SummationByPartsOperators.has_expmv

SummationByPartsOperators.has_expmv!

SummationByPartsOperators.has_ldiv

SummationByPartsOperators.has_ldiv!

SummationByPartsOperators.has_mul

SummationByPartsOperators.has_mul!

SummationByPartsOperators.hermite

SummationByPartsOperators.include

SummationByPartsOperators.includes_boundaries

SummationByPartsOperators.init

SummationByPartsOperators.initialize!

PolynomialBases.integrate — Function.

integrate(func, u, D::AbstractPeriodicDerivativeOperator)

Map the function func to the coefficients u and integrate with respect to the quadrature rule associated with the derivative operator D.

integrate(func, u, D::PeriodicDerivativeOperator)

Map the function func to the coefficients u and integrate with respect to the quadrature rule associated with the periodic derivative operator D.

integrate(func, u, D::DerivativeOperator)

Map the function func to the coefficients u and integrate with respect to the quadrature rule associated with the SBP derivative operator D.

SummationByPartsOperators.interpolate

SummationByPartsOperators.interpolate!

SummationByPartsOperators.interpolation_matrix

SummationByPartsOperators.interpolation_matrix!

SummationByPartsOperators.intervals

SummationByPartsOperators.is_constant

SummationByPartsOperators.isdiscrete

SummationByPartsOperators.isinplace

SummationByPartsOperators.jacobi

SummationByPartsOperators.jacobi_vandermonde

SummationByPartsOperators.laguerre

SummationByPartsOperators.left_boundary_weight

SummationByPartsOperators.left_length

SummationByPartsOperators.legendre

SummationByPartsOperators.legendre_D

SummationByPartsOperators.legendre_M

SummationByPartsOperators.legendre_derivative_operator

SummationByPartsOperators.legendre_vandermonde

SummationByPartsOperators.lower_bandwidth

SummationByPartsOperators.map_from_canonical

SummationByPartsOperators.map_from_canonical!

SummationByPartsOperators.map_to_canonical

SummationByPartsOperators.map_to_canonical!

SummationByPartsOperators.mass_matrix — Function.

mass_matrix(D::Union{DerivativeOperator,VarCoefDerivativeOperator})

Create the diagonal mass matrix for the SBP derivative operator D.

LinearAlgebra.mul! — Function.

mul!(dest::AbstractVector, coefficients::PeriodicDerivativeCoefficients, u::AbstractVector, α, β)

Compute α*D*u + β*dest and store the result in dest.

mul!(dest::AbstractVector, coefficients::PeriodicDerivativeCoefficients, u::AbstractVector, α)

Compute α*D*u and store the result in dest.

mul!(dest::AbstractVector, D::PeriodicDerivativeOperator, u::AbstractVector, α, β)

Compute α*D*u + β*dest and store the result in dest.

mul!(dest::AbstractVector, D::PeriodicDerivativeOperator, u::AbstractVector, α)

Compute α*D*u and store the result in dest.

mul!(dest::AbstractVector, D::PeriodicDissipationOperator, u::AbstractVector, α, β)

Compute α*D*u + β*dest and store the result in dest.

mul!(dest::AbstractVector, D::PeriodicDissipationOperator, u::AbstractVector, α)

Compute α*D*u and store the result in dest.

mul!(dest::AbstractVector, coefficients::DerivativeCoefficients, u::AbstractVector, α, β)

Compute α*D*u + β*dest and store the result in dest.

mul!(dest::AbstractVector, coefficients::DerivativeCoefficients, u::AbstractVector, α)

Compute α*D*u and store the result in dest.

mul!(dest::AbstractVector, D::DerivativeOperator, u::AbstractVector, α, β)

Compute α*D*u + β*dest and store the result in dest.

mul!(dest::AbstractVector, D::DerivativeOperator, u::AbstractVector, α)

Compute α*D*u and store the result in dest.

mul!(dest::AbstractVector, coefficients::VarCoefDerivativeCoefficients, u::AbstractVector, b::AbstractVector, α, β)

Compute α*D*u + β*dest using the coefficients b and store the result in dest.

mul!(dest::AbstractVector, coefficients::VarCoefDerivativeCoefficients, u::AbstractVector, b::AbstractVector, α)

Compute α*D*u using the coefficients b and store the result in dest.

mul!(dest::AbstractVector, D::DissipationOperator, u::AbstractVector, α, β)

Compute α*D*u + β*dest and store the result in dest.

mul!(dest::AbstractVector, D::DissipationOperator, u::AbstractVector, α)

Compute α*D*u and store the result in dest.

mul!(dest::AbstractVector, D::VarCoefDerivativeOperator, u::AbstractVector, α, β)

Compute α*D*u + β*dest and store the result in dest.

mul!(dest::AbstractVector, D::VarCoefDerivativeOperator, u::AbstractVector, α)

Compute α*D*u and store the result in dest.

SummationByPartsOperators.mul!_unrolled_expansion_

SummationByPartsOperators.offset

SummationByPartsOperators.periodic_central_derivative_coefficients — Function.

periodic_central_derivative_coefficients(derivative_order, accuracy_order, T=Float64, parallel=Val{:serial}())

Create the PeriodicDerivativeCoefficients approximating the derivative_order-th derivative with an order of accuracy accuracy_order and scalar type T. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

SummationByPartsOperators.periodic_central_derivative_operator — Function.

periodic_central_derivative_operator(derivative_order, accuracy_order, xmin, xmax, N, parallel=Val{:serial}())

Create a PeriodicDerivativeOperator approximating the derivative_order-th derivative on a uniform grid between xmin and xmax with N grid points up to order of accuracy accuracy_order. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

periodic_central_derivative_operator(derivative_order, accuracy_order, grid, parallel=Val{:serial}())

Create a PeriodicDerivativeOperator approximating the derivative_order-th derivative on the uniform grid up to order of accuracy accuracy_order. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

SummationByPartsOperators.periodic_derivative_coefficients — Function.

function periodic_derivative_coefficients(derivative_order, accuracy_order, left_offset=-(accuracy_order+1)÷2, T=Float64, parallel=Val{:serial}())

Create the PeriodicDerivativeCoefficients approximating the derivative_order-th derivative with an order of accuracy accuracy_order and scalar type T where the leftmost grid point used is determined by left_offset. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}())`.

periodic_derivative_coefficients(source::Holoborodko2008, derivative_order, accuracy_order;
                                 T=Float64, parallel=Val{:serial}(),
                                 stencil_width=accuracy_order+3)

Create the PeriodicDerivativeCoefficients approximating the derivative_order-th derivative with an order of accuracy accuracy_order and scalar type T given by Holoborodko2008. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}())`.

SummationByPartsOperators.periodic_derivative_operator — Function.

periodic_derivative_operator(derivative_order, accuracy_order, xmin, xmax, N, left_offset=-(accuracy_order+1)÷2, parallel=Val{:serial}())

Create a PeriodicDerivativeOperator approximating the derivative_order-th derivative on a uniform grid between xmin and xmax with N grid points up to order of accuracy accuracy_order where the leftmost grid point used is determined by left_offset. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

periodic_derivative_operator(source::Holoborodko2008, derivative_order, accuracy_order,
                             xmin, xmax, N; parallel=Val{:serial}(), kwargs...)

Create a PeriodicDerivativeOperator approximating the derivative_order-th derivative on a uniform grid between xmin and xmax with N grid points up to order of accuracy accuracy_order where the leftmost grid point used is determined by left_offset. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

periodic_derivative_operator(derivative_order, accuracy_order, grid, left_offset=-(accuracy_order+1)÷2, parallel=Val{:serial}())

Create a PeriodicDerivativeOperator approximating the derivative_order-th derivative on thr uniform grid up to order of accuracy accuracy_order where the leftmost grid point used is determined by left_offset. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

SummationByPartsOperators.rand_cache

SummationByPartsOperators.ratenoise_cache

SummationByPartsOperators.reeval_internals_due_to_modification!

SummationByPartsOperators.reinit!

SummationByPartsOperators.remake

SummationByPartsOperators.resize!

SummationByPartsOperators.resize_non_user_cache!

SummationByPartsOperators.right_boundary_weight

SummationByPartsOperators.right_length

SummationByPartsOperators.satisfies_sbp

SummationByPartsOperators.savevalues!

SummationByPartsOperators.second_derivative_coefficients

SummationByPartsOperators.semidiscretise — Function.

semidiscretise(u0func, semidisc::AbstractSemidiscretisation, tspan)

Apply the semidiscretisation semidisc to the initial data given by u0func and return an ODEProblem with time span tspan.

SummationByPartsOperators.set_abstol!

SummationByPartsOperators.set_filter_coefficients!

SummationByPartsOperators.set_proposed_dt!

SummationByPartsOperators.set_reltol!

SummationByPartsOperators.set_t!

SummationByPartsOperators.set_u!

SummationByPartsOperators.solve

SummationByPartsOperators.solve!

SummationByPartsOperators.source_of_coeffcients

SummationByPartsOperators.source_of_coefficients

SummationByPartsOperators.spectral_viscosity_operator

SummationByPartsOperators.step!

SummationByPartsOperators.super_spectral_viscosity_operator

SummationByPartsOperators.terminate!

SummationByPartsOperators.third_derivative_coefficients

SummationByPartsOperators.tuples

SummationByPartsOperators.u_cache

SummationByPartsOperators.u_modified!

SummationByPartsOperators.update_coefficients

SummationByPartsOperators.update_coefficients!

SummationByPartsOperators.upper_bandwidth

SummationByPartsOperators.user_cache

SummationByPartsOperators.utility_matrices

SummationByPartsOperators.var_coef_derivative_coefficients

SummationByPartsOperators.var_coef_derivative_operator — Function.

var_coef_derivative_operator(source_of_coefficients, derivative_order, accuracy_order, xmin, xmax, N, left_weights, right_weights, bfunc, parallel=Val{:serial}())

Create a VarCoefDerivativeOperator approximating a derivative_order-th derivative with variable coefficients bfunc on a grid between xmin and xmax with N grid points up to order of accuracy accuracy_order with coefficients given by source_of_coefficients. The evaluation of the derivative can be parallised using threads by chosing parallel=Val{:threads}()).

SummationByPartsOperators.warn_compat