Docstrings

StaticGraphs.AbstractStaticEdge

StaticGraphs.AbstractStaticGraph — Type.

AbstractStaticGraph{T, U}

An abstract type representing a simple graph structure parameterized by integer types

T: the type representing the graph's vertices

U: the type representing the number of edges in the graph

StaticGraphs.SDGFormat

StaticGraphs.SDiGraph

StaticGraphs.SGFormat

StaticGraphs.SGraph

StaticGraphs.StaticDiGraph — Type.

StaticDiGraph{T, U} <: AbstractStaticGraph{T, U}

A type representing a static directed graph.

StaticGraphs.StaticDiGraphEdge

StaticGraphs.StaticEdge

StaticGraphs.StaticEdgeIter

StaticGraphs.StaticGraph — Type.

StaticGraph{T, U} <: AbstractStaticGraph{T, U}

A type representing an undirected static graph.

StaticGraphs.StaticGraphEdge

StaticGraphs.StaticGraphFormat

StaticGraphs.StaticGraphs

StaticGraphs.UnsafeVectorView

StaticGraphs._bvrange

StaticGraphs._fvrange

StaticGraphs.eval

StaticGraphs.fastview — Function.

UnsafeVectorView only works for isbits types. For other types, we're already allocating lots of memory elsewhere, so creating a new SubArray is fine. This function looks type-unstable, but the isbits(T) test can be evaluated by the compiler, so the result is actually type-stable. From https://github.com/rdeits/NNLS.jl/blob/0a9bf56774595b5735bc738723bd3cb94138c5bd/src/NNLS.jl#L218.

StaticGraphs.in_edges

StaticGraphs.include

StaticGraphs.indtype

StaticGraphs.loadsg

StaticGraphs.mintype — Function.

mintype(v)

Returns the minimum integer type required to represent integer v.

StaticGraphs.out_edges

StaticGraphs.savesg

StaticGraphs.vectype