Docstrings

MarriageMarkets.MarriageMarkets — Module.

A package for solving various marriage market matching models.

Two models are supported:

StaticMatch: the static frictionless matching model of Choo & Siow (2006)

SearchMatch: the search and matching model of Shimer & Smith (2000)

SearchMatch is generalized in a few ways beyond the model presented in the paper. This includes a match-specific "love" shock to make the matching probabilistic.

MarriageMarkets.STDNORMAL

MarriageMarkets.SearchClosed — Function.

SearchClosed(λ, δ, r, σ, θ_m, θ_f, ℓ_m, ℓ_f, g; β=0.5, verbose=false, step=0.2)

Constructs marriage market equilibrium of closed-system model with match-specific productivity shocks and production function $g(x,y)$.

SearchClosed(λ, δ, r, σ, θ_m, θ_f, ℓ_m, ℓ_f, g; β=0.5, verbose=false, step=0.2)

Method for one-dimensional types.

MarriageMarkets.SearchInflow — Function.

SearchInflow(λ, δ, r, σ, θ_m, θ_f, γ_m, γ_f, ψ_m, ψ_f, g; β=0.5, verbose=false, step=0.2)

Constructs marriage market equilibrium of inflow and death model with match-specific productivity shocks and production function $g(x,y)$.

SearchInflow(λ, δ, r, σ, θ_m, θ_f, γ_m, γ_f, ψ_m, ψ_f, g; β=0.5, verbose=false, step=0.2)

Method for one-dimensional types.

MarriageMarkets.SearchMatch — Type.

SearchMatch(λ, δ, r, σ, θ_m, θ_f, γ_m, γ_f, ψ_m, ψ_f, ℓ_m, ℓ_f, h; β=0.5, verbose=false, step=0.2)

Construct a Shimer & Smith (2000) marriage market model and solve for the equilibrium. When match-specific shocks are included, the divorce process is endogenized as in Goussé (2014).

The equilibrium is the solution to a nested fixed-point mapping. The inner fixed-point solves for a matching equilibrium: consistent strategies given fixed singles densities. The outer fixed-point solves for a market equilibrium: singles densities that are consistent with strategies.

Model selection depends which arguments are provided:

Match-specific additive shocks $z ~ N(0, σ)$ when $σ > 0$

* Note: some randomness is required for the fixed point iteration to converge with sex asymmetry

Closed system or inflow/outflow:

* `ℓ_m, ℓ_f` exogenous: population circulates between singlehood and marriage, no birth/death
* Death rates `ψ_m, ψ_f`, inflows `γ_m, γ_f`: population distributions `ℓ_m, ℓ_f` endogenous

MarriageMarkets.StaticMatch — Type.

StaticMatch(mtypes, ftypes, mdist, fdist, surplus)

Construct a Choo & Siow (2006) marriage market model and solve for the equilibrium.

MarriageMarkets.demand — Function.

Quasi-demand functions

Assuming that surplus is generated from equal gains on both sides. This is w.l.o.g. as gains and transfers cannot be separately identified.

MarriageMarkets.equilibrium — Function.

Compute equilibrium shares of singles and marriages.

Solves the matching equilibrium by searching for the transfers (analogous to prices) which clear the market.

MarriageMarkets.estimate_static_surplus — Function.

Estimate the marital surplus from observed matches and singles.

MarriageMarkets.eval

MarriageMarkets.generate_surplus — Function.

Construct production array from function.

MarriageMarkets.include

MarriageMarkets.prod_array — Function.

Construct production array from function.

MarriageMarkets.sex_solve — Function.

Wrapper for nlsolve that handles the concatenation and splitting of sex-specific arrays.

MarriageMarkets.sex_split — Function.

Split vector v into male/female pieces, where idx is number of male types.

MarriageMarkets.surplusdiv — Function.

Wife's consumption out of surplus (aggregate share).