Docstrings

ConditionalJuMP.@?

ConditionalJuMP.@disjunction

ConditionalJuMP.@ifelse

ConditionalJuMP.@implies

ConditionalJuMP.@switch

ConditionalJuMP.Conditional

ConditionalJuMP.ConditionalJuMP

ConditionalJuMP.Constraint

ConditionalJuMP.Implication

ConditionalJuMP.Indicator

ConditionalJuMP.IndicatorMap

ConditionalJuMP.NArg

ConditionalJuMP.UnboundedVariableException

ConditionalJuMP.UnhandledComplementException

ConditionalJuMP._all_conditional

ConditionalJuMP._conditional

ConditionalJuMP._conditionalize

ConditionalJuMP._conditionalize_function_broadcast

ConditionalJuMP._conditionalize_operator_broadcast

ConditionalJuMP._fix

ConditionalJuMP._getvalue — Function.

Like JuMP.getvalue, but returns a Nullable{T}() for unset variables instead of throwing a warning

ConditionalJuMP._hash

ConditionalJuMP._setvalue

ConditionalJuMP._unfix

ConditionalJuMP.append

ConditionalJuMP.complement

ConditionalJuMP.disjunction!

ConditionalJuMP.eval

ConditionalJuMP.getindicator!

ConditionalJuMP.getindmap!

ConditionalJuMP.getmodel

ConditionalJuMP.handle_constant_objective! — Function.

Gurobi doesn't understand constant terms in the objective, which can affect the way the relative MIP gap is interpreted. To work around this, we'll replace the constant term k with an affine term 1*x for a new variable x fixed to equal k.

ConditionalJuMP.hascomplement

ConditionalJuMP.ifelse_conditional

ConditionalJuMP.implies!

ConditionalJuMP.include

ConditionalJuMP.interval

ConditionalJuMP.is_function_broadcast

ConditionalJuMP.is_operator_broadcast

ConditionalJuMP.isfixed

ConditionalJuMP.isjump

ConditionalJuMP.lowerbound

ConditionalJuMP.nan_to_null

ConditionalJuMP.newbinaryvar

ConditionalJuMP.newcontinuousvar

ConditionalJuMP.normalize!

ConditionalJuMP.simplify! — Function.

Simplification function that chooses the appropriate algorithm based on the number of variables.

ConditionalJuMP.simplify_dict! — Function.

O(N) simplification, but with a substantially larger constant cost due to the need to construct a Dict.

ConditionalJuMP.simplify_inplace! — Function.

Naive O(N^2) simplification. Slower for very large expressions, but allocates no memory and is solidly faster for expressions with < 100 variables.

ConditionalJuMP.switch

ConditionalJuMP.switch!

ConditionalJuMP.upperbound

ConditionalJuMP.variable

ConditionalJuMP.warmstart!