Docstrings

InPlace.@inplace — Macro.

@inplace a = f(args...)
@inplace a += expr

Compute f(args...) resp. +(a, expr) and assign the result to a. If a is mutable, possibly modify its value in-place.

In the case where a is mutable, it is an implementation detail whether its value is actually modified. For this reason, one should use this operation only on values for which the current stackframe holds the only reference; e.g. by using deepcopy.

InPlace.InPlace

InPlace.eval

InPlace.include

InPlace.inclusiveinplace! — Function.

a = inclusiveinplace!(op, a, b...)

Compute and return op(a, b...). If a is mutable, possibly modify its value in-place.

In the case where a is mutable, it is an implementation detail whether its value is actually modified, and for this reason, one should always also assign the result of this call to a. Moreover, one should use it only on values for which the current stackframe holds the only reference; e.g. by using deepcopy.

InPlace.inplace! — Function.

a = inplace!(op, a, b...)

Compute and return op(b...). If a is mutable, possibly modify its value in-place.

In the case where a is mutable, it is an implementation detail whether its value is actually modified, and for this reason, one should always also assign the result of this call to a. Moreover, one should use it only on values for which the current stackframe holds the only reference; e.g. by using deepcopy.