Docstrings

Nullables.BINARY_OPS

Nullables.EqualOrLess

Nullables.NullException

Nullables.NullSafeFloats

Nullables.NullSafeInts

Nullables.NullSafeSignedInts

Nullables.NullSafeTypes

Nullables.NullSafeUnsignedInts

Nullables.Nullable — Type.

Nullable(x, hasvalue::Bool=true)

Wrap value x in an object of type Nullable, which indicates whether a value is present. Nullable(x) yields a non-empty wrapper and Nullable{T}() yields an empty instance of a wrapper that might contain a value of type T.

Nullable(x, false) yields Nullable{typeof(x)}() with x stored in the result's value field.

Examples

julia> Nullable(1)
Nullable{Int64}(1)

julia> Nullable{Int64}()
Nullable{Int64}()

julia> Nullable(1, false)
Nullable{Int64}()

julia> dump(Nullable(1, false))
Nullable{Int64}
  hasvalue: Bool false
  value: Int64 1

Nullables.Nullables

Nullables._nullable_eltype

Nullables._unsafe_get_eltype

Nullables.broadcast_axes

Nullables.eval

Nullables.hasvalue

Nullables.include

Nullables.isnull — Function.

isnull(x::Nullable)

Return whether or not x is null.

Examples

julia> x = Nullable(1, false)
Nullable{Int64}()

julia> isnull(x)
true

julia> x = Nullable(1, true)
Nullable{Int64}(1)

julia> isnull(x)
false

julia> x = 1
1

julia> isnull(x)
false

Nullables.maptoTuple

Nullables.null_safe_op — Function.

null_safe_op(f::Any, ::Type, ::Type...)::Bool

Returns whether an operation f can safely be applied to any value of the passed type(s). Returns false by default.

Custom types should implement methods for some or all operations f when applicable: returning true means that the operation may be called on any bit pattern without throwing an error (though returning invalid or nonsensical results is not a problem). In particular, this means that the operation can be applied on the whole domain of the type and on uninitialized objects. As a general rule, these properties are only true for safe operations where isbitstype is true.

Types declared as safe can benefit from higher performance for operations on nullable: by always computing the result even for null values, a branch is avoided, which helps vectorization.

Nullables.nullable_returntype — Function.

Return the given type if it is concrete, and Union{} otherwise.

Nullables.nullablehash_seed

Nullables.unsafe_get — Function.

unsafe_get(x)

Return the value of x for Nullable x; return x for all other x.

This method does not check whether or not x is null before attempting to access the value of x for x::Nullable (hence "unsafe").

Examples

julia> x = Nullable(1)
Nullable{Int64}(1)

julia> unsafe_get(x)
1

julia> x = Nullable{String}()
Nullable{String}()

julia> unsafe_get(x)
ERROR: UndefRefError: access to undefined reference
Stacktrace:
 [1] unsafe_get(::Nullable{String}) at ./nullable.jl:152

julia> x = 1
1

julia> unsafe_get(x)
1