Docstrings

GLMakie.@csafe — Macro.

Throwing an error in a c callback seems to lead to undefined behaviour

GLMakie.AbstractContext

GLMakie.GLAbstraction

GLMakie.GLBackend

GLMakie.GLContext

GLMakie.GLFramebuffer

GLMakie.GLMakie

GLMakie.GLScreen

GLMakie.GLVisualize

GLMakie.MonitorProperties

GLMakie.PostProcessROBJ

GLMakie.PostprocessPrerender

GLMakie.Screen

GLMakie.ScreenArea

GLMakie.ScreenID

GLMakie.SelectionID — Type.

Selection of random objects on the screen is realized by rendering an object id + plus an arbitrary index into the framebuffer. The index can be used for e.g. instanced geometries.

GLMakie.WORKING_OPENGL

GLMakie.ZIndex

GLMakie.__init__

GLMakie._global_gl_screen

GLMakie._mouse_selection_id

GLMakie.addbuttons

GLMakie.assetpath — Function.

returns path relative to the assets folder

GLMakie.atlas_texture_cache

GLMakie.attach_framebuffer

GLMakie.cached_robj!

GLMakie.convert_mesh_color

GLMakie.correct_mouse

GLMakie.destroy!

GLMakie.draw_atomic

GLMakie.draw_fullscreen

GLMakie.eval

GLMakie.fast_color_data!

GLMakie.flatten_plots

GLMakie.framebuffer_size

GLMakie.get_image

GLMakie.get_texture!

GLMakie.getscreen

GLMakie.gl_screens

GLMakie.global_gl_screen

GLMakie.gpuvec

GLMakie.handle_intensities!

GLMakie.handle_view

GLMakie.id2rect

GLMakie.include

GLMakie.index1D

GLMakie.lift_convert

GLMakie.loadasset — Function.

Loads a file from the asset folder

GLMakie.loadshader

GLMakie.make_context_current

GLMakie.makieshader

GLMakie.mouse_selection_native

GLMakie.opengl_renderloop

GLMakie.pick_native

GLMakie.pixel2world

GLMakie.postprocess — Function.

Creates a postprocessing render object. This will transfer the pixels from the color texture of the Framebuffer to the screen and while at it, it can do some postprocessing (not doing it right now): E.g fxaa anti aliasing, color correction etc.

GLMakie.rcpframe

GLMakie.reactive_run_till_now

GLMakie.remove_automatic!

GLMakie.render_frame — Function.

Renders a single frame of a window

GLMakie.renderloop

GLMakie.retina_scaling_factor

GLMakie.rewrap — Function.

OpenGL shares all data containers between shared contexts, but not vertexarrays -.- So to share a robjs between a context, we need to rewrap the vertexarray into a new one for that specific context.

GLMakie.scene2image — Function.

scene2image(scene::Scene)

Buffers the scene in an image buffer.

GLMakie.selection_queries

GLMakie.setup!

GLMakie.sleep_pessimistic — Function.

Sleep is pretty imprecise. E.g. anything under 0.001s is not guaranteed to wake up before 0.001s. So this timer is pessimistic in the way, that it will never sleep more than time.

GLMakie.surface_contours

GLMakie.to_gl_text

GLMakie.to_glvisualize_key

GLMakie.to_jl_layout!

GLMakie.to_native

GLMakie.to_range

GLMakie.to_width

GLMakie.vec2color

GLMakie.volume_prerender

GLMakie.was_destroyed

GLMakie.window_position

GLMakie.window_size