Docstrings

ApproxManifoldProducts.*

ApproxManifoldProducts.A_invB

ApproxManifoldProducts.AngleAxis

ApproxManifoldProducts.ApproxManifoldProducts

ApproxManifoldProducts.AxisAngle

ApproxManifoldProducts.BallTree

ApproxManifoldProducts.BallTreeDensity

ApproxManifoldProducts.CTs

ApproxManifoldProducts.Circular

ApproxManifoldProducts.Euclid2

ApproxManifoldProducts.EuclideanManifold

ApproxManifoldProducts.Euler

ApproxManifoldProducts.FloatInt

ApproxManifoldProducts.KDE

ApproxManifoldProducts.KernelDensityEstimate

ApproxManifoldProducts.MB

ApproxManifoldProducts.Manifold

ApproxManifoldProducts.ManifoldBelief

ApproxManifoldProducts.MixtureDensity

ApproxManifoldProducts.Ndim

ApproxManifoldProducts.Npts

ApproxManifoldProducts.Quaternion

ApproxManifoldProducts.R

ApproxManifoldProducts.SE2

ApproxManifoldProducts.SE3

ApproxManifoldProducts.SO3

ApproxManifoldProducts.TUs

ApproxManifoldProducts.TransformUtils

ApproxManifoldProducts.VectorFloatInt

ApproxManifoldProducts.VectorRange

ApproxManifoldProducts.\

ApproxManifoldProducts.__init__

ApproxManifoldProducts.addtheta

ApproxManifoldProducts.buildHybridManifoldCallbacks

ApproxManifoldProducts.compare

ApproxManifoldProducts.convert

ApproxManifoldProducts.convert!

ApproxManifoldProducts.deltaso3vee

ApproxManifoldProducts.difftheta

ApproxManifoldProducts.doinversetransform!

ApproxManifoldProducts.eval

ApproxManifoldProducts.evalAvgLogL

ApproxManifoldProducts.evaluateDualTree

ApproxManifoldProducts.evaluateManifoldNaive1D! — Function.

evaluateManifoldNaive1D!(ret, idx, pts, bw, x)
evaluateManifoldNaive1D!(ret, idx, pts, bw, x, loo)
evaluateManifoldNaive1D!(ret, idx, pts, bw, x, loo, diffop)

Evalute the KDE naively as equally weighted Gaussian kernels with common bandwidth. This function does, however, allow on-manifold evaluations.

ApproxManifoldProducts.expmOwn

ApproxManifoldProducts.expmOwn1

ApproxManifoldProducts.expmOwn2

ApproxManifoldProducts.expmOwn3

ApproxManifoldProducts.expmOwn4

ApproxManifoldProducts.expmOwnT

ApproxManifoldProducts.get2DLambda

ApproxManifoldProducts.get2DMu

ApproxManifoldProducts.get2DMuMin

ApproxManifoldProducts.getBW

ApproxManifoldProducts.getCircLambda

ApproxManifoldProducts.getCircMu

ApproxManifoldProducts.getKDEManifoldBandwidths

ApproxManifoldProducts.getKDEMax

ApproxManifoldProducts.getKDEMean

ApproxManifoldProducts.getKDERange

ApproxManifoldProducts.getKDEfit

ApproxManifoldProducts.getPoints

ApproxManifoldProducts.getWeights

ApproxManifoldProducts.include

ApproxManifoldProducts.intersIntgAppxIS

ApproxManifoldProducts.inverse

ApproxManifoldProducts.kde!

ApproxManifoldProducts.kde!_CircularNaiveCV

ApproxManifoldProducts.kld

ApproxManifoldProducts.logmap

ApproxManifoldProducts.logmap_SO2

ApproxManifoldProducts.manifoldLooCrossValidation — Function.

manifoldLooCrossValidation(pts, bw; own, diffop)

Calculate negative entropy with leave one out (j'th element) cross validation.

Background

From: Silverman, B.: Density Estimation for Statistics and Data Analysis, 1986, p.52

Probability density function p(x), as estimated by kernels

\[hatp_{-j}(x) = 1/(N-1) Σ_{i != j}^N frac{1}{sqrt{2pi}σ } exp{ -frac{(x-μ)^2}{2 σ^2} }\]

and has Cross Validation number as the average log evaluations of leave one out hatp_{-j}(x):

\[CV(p) = 1/N Σ_i^N log hat{p}_{-j}(x_i)\]

This quantity CV is related to an entropy H(p) estimate via:

\[H(p) = -CV(p)\]

ApproxManifoldProducts.manifoldProduct — Function.

manifoldProduct(ff, manif; Niter)

Approximate the pointwise the product of functionals on manifolds using KernelDensityEstimate.

Example

using ApproxManifoldProducts

# two densities on a cylinder
p = manikde!(randn(2,100), (:Euclid, :Circular) )

pts2a = 3.0*randn(1,100).+5.0
pts2b = TransformUtils.wrapRad.(0.5*randn(1,100).+pi)
q = manikde!([pts2a;pts2b], (:Euclid, :Circular) )

# approximate the product between hybrid manifold densities
pq = manifoldProduct([p;q], (:Euclid, :Circular))

# convenient plotting (work in progress...)
# TODO update docs

# direct histogram plot
using Gadfly
plot( x=getPoints(pq)[1,:], y=getPoints(pq)[2,:], Geom.histogram2d )

ApproxManifoldProducts.manikde! — Function.

manikde!(pts, bws, manifolds)

Legacy extension of KDE.kde! function to approximate smooth functions based on samples, using likelihood cross validation for bandwidth selection. This method allows approximation over hybrid manifolds.

ApproxManifoldProducts.marginal

ApproxManifoldProducts.matrix

ApproxManifoldProducts.matrix!

ApproxManifoldProducts.normDistAccAt! — Function.

normDistAccAt!(ret, idx, x, sigma)
normDistAccAt!(ret, idx, x, sigma, w)

Probability density function p(x), as estimated by kernels

\[hatp_{-j}(x) = 1/(N-1) Σ_{i != j}^N frac{1}{sqrt{2pi}σ } exp{ -frac{(x-μ)^2}{2 σ^2} }\]

ApproxManifoldProducts.normalize

ApproxManifoldProducts.normalize!

ApproxManifoldProducts.ominus

ApproxManifoldProducts.oplus

ApproxManifoldProducts.prodAppxMSGibbsS

ApproxManifoldProducts.q_conj

ApproxManifoldProducts.q_conj!

ApproxManifoldProducts.rand

ApproxManifoldProducts.rbf

ApproxManifoldProducts.rbf!

ApproxManifoldProducts.rbfAccAt!

ApproxManifoldProducts.reci_s2pi

ApproxManifoldProducts.resample

ApproxManifoldProducts.resid2DLinear

ApproxManifoldProducts.rightJacExmap

ApproxManifoldProducts.rightJacExmapinv

ApproxManifoldProducts.root

ApproxManifoldProducts.rotate

ApproxManifoldProducts.rotate!

ApproxManifoldProducts.sample

ApproxManifoldProducts.se2vee

ApproxManifoldProducts.se2vee!

ApproxManifoldProducts.setForceEvalDirect!

ApproxManifoldProducts.setso3!

ApproxManifoldProducts.skew

ApproxManifoldProducts.so3

ApproxManifoldProducts.solveresid2DLinear

ApproxManifoldProducts.solveresid2DLinear!

ApproxManifoldProducts.string

ApproxManifoldProducts.transpose

ApproxManifoldProducts.vec

ApproxManifoldProducts.vee

ApproxManifoldProducts.vee!

ApproxManifoldProducts.veeAngleAxis

ApproxManifoldProducts.veeEuler

ApproxManifoldProducts.veeQuaternion

ApproxManifoldProducts.wrapRad

ApproxManifoldProducts.⊕

ApproxManifoldProducts.⊖