Docstrings

ForceImport.@force — Macro.

    @force using Module

Forces imports of exported methods from Module, even if there are conflicts.

ForceImport.@merge — Macro.

    @merge fun(x) = x

Automatically imports fun from its module if necessary for function definition.

ForceImport.@port — Macro.

    ForceImport.@port using Module

Force exports methods from Module in current module.

ForceImport.ForceImport

ForceImport.eval

ForceImport.imp

ForceImport.include