Docstrings

FieldMetadata.@chain — Macro.

Chain together any macros. Useful for combining @metadata macros.

Example

@chain columns @label @units @default_kw

@columns struct Foo
  bar::Int | 7 | u"g" | "grams of bar"
end

FieldMetadata.@default

FieldMetadata.@description

FieldMetadata.@flattenable

FieldMetadata.@label

FieldMetadata.@limits

FieldMetadata.@metadata — Macro.

Generate a macro that constructs methods of the same name. These methods return the metadata information provided for each field of the struct.

@metadata def_range (0, 0)
@def_range struct Model
    a::Int | (1, 4)
    b::Int | (4, 9)
end

model = Model(3, 5)
def_range(model, Val{:a})
(1, 4)

def_range(model)
((1, 4), (4, 9))

FieldMetadata.@plottable

FieldMetadata.@prior

FieldMetadata.@redefault

FieldMetadata.@redescription

FieldMetadata.@reflattenable

FieldMetadata.@relabel

FieldMetadata.@relimits

FieldMetadata.@replottable

FieldMetadata.@reprior

FieldMetadata.@reunits

FieldMetadata.@units

FieldMetadata.FieldMetadata

FieldMetadata.add_field_funcs

FieldMetadata.addmethod!

FieldMetadata.addmethod_unlessdefault!

FieldMetadata.chained_macros

FieldMetadata.chained_macros!

FieldMetadata.default

FieldMetadata.description

FieldMetadata.eval

FieldMetadata.fieldname_vals

FieldMetadata.firsthead

FieldMetadata.flattenable

FieldMetadata.getkey

FieldMetadata.include

FieldMetadata.label

FieldMetadata.limits

FieldMetadata.namify

FieldMetadata.plottable

FieldMetadata.prior

FieldMetadata.units