Docstrings

SimplePosets.Antichain — Function.

AntiChain(n) creates a new poset with Int elements 1:n with no relations between those elements.

AntiChain(list) creates a new poset with elements from list with no relations between those elements.

SimplePosets.BooleanLattice — Function.

BooleanLattice(n) creates the Boolean lattice whose elements are n-long character strings of 0s and 1s. Ordering is coordinatewise.

SimplePosets.Chain — Function.

Chain(n) creates a new poset whose elements are the Int values 1:n with the relations 1<2<3<...<n.

Chain(list) creates a new poset whose elements are the values in list with list[1]<list[2]<....

SimplePosets.ComparabilityGraph — Function.

ComparabilityGraph(P) returns the comparability graph of the poset P. The result is a SimpleGraph whose vertices are the elements of P in which there is an edge (u,v) provided (u,v) or (v,u) is a relation of P.

SimplePosets.CoverDigraph — Function.

CoverDigraph(P) creates a SimpleDirectedGraph G whose vertices are the elements of P and in which we have the edge (u,v) provided: (a) u<v is a relation of P and (b) there is no w in P with u<w<v.

SimplePosets.Divisors — Function.

Divisors(n) creates a new poset whose elements are the positive divisor of n (including 1 and n itself) in which we have the relations (u,v) precisely when u is a factor of v.

SimplePosets.IntPoset — Function.

IntPoset(n) creates a new SimplePoset{Int}() prepopulated with elements 1:n but no relations.

SimplePosets.PartitionLattice — Function.

PartitionLattice(n) creates the poset whose elements are the partitions of the set {1,2,...,n} ordered by refinement.

SimplePosets.RandomPoset — Function.

RandomPoset(n,d) creates a random d-dimensional poset with elements 1:n.

SimplePosets.SimplePoset — Type.

SimplePoset() creates a new partially ordered set (poset) in which the elements can be of Any type.

Use SimplePoset(T) or SimplePlot{T}() to create a new poset in which the elements are of type T.

SimplePosets.SimplePosets

SimplePosets.StandardExample — Function.

StandardExample(n) creates a new poset with 2n elements in two levels. Each element on the lower level is below exactly n-1 elements from the upper level.

We name the lower elements -1,-2,...,-n and the upper elements 1,2,...,n. We have -i below all the positive elements except +i.

julia> P = StandardExample(4)
SimplePoset{Int64} (8 elements)

julia> elements(P)
8-element Array{Int64,1}:
 -4
 -3
 -2
 -1
  1
  2
  3
  4

julia> relations(P)
12-element Array{Tuple{Int64,Int64},1}:
 (-4,1)
 (-4,2)
 (-4,3)
 (-3,1)
 (-3,2)
 (-3,4)
 (-2,1)
 (-2,3)
 (-2,4)
 (-1,2)
 (-1,3)
 (-1,4)

SimplePosets._LX_table

SimplePosets.above — Function.

above(P,x) returns a list of all elements y in the poset P for which x<y.

SimpleGraphs.add! — Function.

add!(P,x) adds the element x to the poset with no relations to any other elements.

add!(P,x,y) adds elements x and y to the poset (if they are not already present) and, more importantly, adds the relation x<y as well. This fails if adding this relation would violate transitivity.

SimplePosets.all_linear_extensions — Function.

all_linear_extensions(P) returns the Set of all linear extensions of P. This can take a very long time and eat up a lot of memory (which can be freed using clear_LX_table).

SimplePosets.below — Function.

below(P,x) returns a list of all elements y in the poset P for which y<x.

SimplePosets.card — Function.

card(P) returns the cardinality (number of elements) of P.

SimplePosets.check

SimplePosets.clear_LX_table — Function.

clear_LX_table() releases cached results computed by all_linear_extensions.

Base.delete! — Function.

delete!(P,x) deletes the element x from the poset P.

delete!(P,x,y) deletes the relation x<y from P (assuming that such a relation exists in P). It then deletes other relations as needed to ensure what remains is still a poset (i.e., still transitive).

SimplePosets.display

SimplePosets.divisors

SimplePosets.element_type — Function.

element_type(P) returns the type of elements in this SimplePoset.

SimplePosets.elements — Function.

elements(P) returns a list of the elements of P.

SimplePosets.eval

SimplePosets.first_prime_factor

SimpleGraphs.has — Function.

has(P,x) checks if x is an element of the poset P.

has(P,x,y) checks if x<y is a relation in the poset P.

SimplePosets.hash

SimplePosets.height — Function.

height(P) gives the size of a largest chain in the poset P.

SimplePosets.include

SimplePosets.incomparables — Function.

incomparables(P) returns a list of ordered pairs (u,v) such that u and v are incomparable in P. Note that if (u,v) appears in the list, we do not also include (v,u).

Base.intersect — Function.

intersect(P,Q) constructs a new poset that is the intersection of the two given posets (which must contain elements of the same datatype).

SimplePosets.interval — Function.

interval(P,x,y) returns a list of all elements z in the poset P for which x<z<y.

Base.inv — Function.

inv(P) creates a new poset with the same elements as P in which all of P's relations have been reversed. That is (u,v) is a relation of P iff (v,u) is a relation of inv(P).

Use P' as a shortcut for inv(P).

SimplePosets.linear_extension — Function.

linear_extension(P) returns a linear extension of the poset P.

SimplePosets.maximals — Function.

maximals(P) returns a list of maximal elements of P.

SimplePosets.minimals — Function.

minimals(P) returns a list of minimal elements of P.

SimplePosets.mobius — Function.

mobius(P) returns the Mobius function of the poset P as a Dict. See zeta.

SimplePosets.mobius_matrix — Function.

mobius_matrix(P) returns the inverse of zeta_matrix(P).

SimplePosets.random_average_height — Function.

random_average_height(P,reps=100) gives the average height of the elements of P in a random linear extension. See random_linear_extension.

SimplePosets.random_linear_extension — Function.

random_linear_extension(P) generates a linear extension of P at random. This is slower than linear_extension(P) if one just wants some linear extension.

Note that linear extensions are not generated uniformly at random by this function. Rather we proceed recursively: A minimal element x of P is chosen uniformly at random from among all minimal elements, and that becomes the first (lowest) element in the linear extension. Then x is deleted and the same process is repeated until all elements have been processed.

SimpleGraphs.relabel — Function.

relabel(P) creates a new poset P in which the elements are relabeled using the values 1:n (where n=card(P)).

relabel(P,d) relabels elements according to the dictionary d.

SimplePosets.relations — Function.

relations(P) returns a list of ordered pairs (u,v) where u and v are elements of the poset that satisfy u<v in P.

SimplePosets.show

SimplePosets.stack — Function.

stack(P...) stacks its poset arguments each on top of each other. The first poset in the list is on the bottom.

SimplePosets.true_average_height — Function.

true_average_height(P) gives the average height of the elements of P averaged over all linear extensions of P. See all_linear_extensions.

SimplePosets.vec_less

SimplePosets.zeta — Function.

zeta(P) returns the zeta function of P as a Dict. If d=zeta(P) then d[u,v]==1 when u and v are elements of the poset for which u==v or (u,v) is a relation of P.

SimplePosets.zeta_matrix — Function.

zeta_matrix(P) creates the zeta matrix of a poset P. The rows/columns of this matrix are indexed by the elements of P. We have a 1 in the i,j-position exactly when the ith element of P is <= the jth element of P.

The ordering of the rows/columns does not necessarily lead to an upper triangular matrix.