Docstrings

TwentyFour.DEFAULT_GOAL — Constant.

The default goal of a Twenty Four problem is 24.

TwentyFour.OPS

TwentyFour.OP_SYMS

TwentyFour.TwentyFour

TwentyFour.XXIV — Type.

XXIV is an instance of a Twenty Four problem.

TwentyFour.eval

TwentyFour.get_goal — Function.

get_goal(X::XXIV) returns the targe goal of this instance of a Twenty Four problem. This is typically 24.

TwentyFour.include

TwentyFour.set_goal — Function.

set_goal(X::XXIV, g) sets the goal of this Twenty Four problem to g. Default g is 24.

TwentyFour.solve — Function.

solve(nums...) solves a Twenty-Four puzzle for the given numbers. Returns a String giving the solution. Example:

julia> solve(4,2,5,1)
"(4+2)*(5-1)"

TwentyFour.solver

TwentyFour.str — Function.

str(x::Rational{Int}) returns a nice string form of x.