Docstrings

ArrayAllez.ARVector

ArrayAllez.ArrayAllez

ArrayAllez.Array_ — Type.

similar_(name, A)     ≈ similar(A)
Array_{T}(name, size) ≈ Array{T}(undef, size)

New arrays for intermediate results, drawn from an LRU cache when length(A) >= 2000. The cache's key uses name::Symbol as well as type & size to ensure different uses don't collide.

copy_(name, A) = copyto!(similar_(name, A), A)

Just like that.

ArrayAllez.CFloat

ArrayAllez.CFloatArray

ArrayAllez.CFloatMatrix

ArrayAllez.CacheKey

ArrayAllez.IVERBOSE

ArrayAllez.RVector

ArrayAllez.VEC

ArrayAllez.__init__

ArrayAllez.__inits__

ArrayAllez._bsum

ArrayAllez.broadsum — Function.

broadsum(*, A, B, C)       = sum(A .* B .* C)
broadsum(f,*, A,B)         = sum(f, A .* B)

These aim to work exactly like sum(broadcast(...)), but without materialising the broadcast array. Simplest case works & is fast. Version with f is slow.

broadsum(*, A, B; dims=1)  = sum(A .* B; dims=1)
broadsum!(Z, *,A,B)        = sum!(Z, A .* B)

Similarly immitates sum!(Z, broadcast(...)) without materialising. Now uses LazyArrays.BroadcastArray. The in-place form is actually a little slower.

ArrayAllez.broadsum!

ArrayAllez.bsum — Function.

broadsum(*, A, B, C)       = sum(A .* B .* C)
broadsum(f,*, A,B)         = sum(f, A .* B)

These aim to work exactly like sum(broadcast(...)), but without materialising the broadcast array. Simplest case works & is fast. Version with f is slow.

broadsum(*, A, B; dims=1)  = sum(A .* B; dims=1)
broadsum!(Z, *,A,B)        = sum!(Z, A .* B)

Similarly immitates sum!(Z, broadcast(...)) without materialising. Now uses LazyArrays.BroadcastArray. The in-place form is actually a little slower.

ArrayAllez.bsum!

ArrayAllez.cache

ArrayAllez.checksum

ArrayAllez.copy_ — Function.

similar_(name, A)     ≈ similar(A)
Array_{T}(name, size) ≈ Array{T}(undef, size)

New arrays for intermediate results, drawn from an LRU cache when length(A) >= 2000. The cache's key uses name::Symbol as well as type & size to ensure different uses don't collide.

copy_(name, A) = copyto!(similar_(name, A), A)

Just like that.

ArrayAllez.eval

ArrayAllez.exp! — Function.

exp!(A)
exp_(A) = exp!(similar(A), A)
exp0(A) ≈ exp.(A)

Element-wise in-place exponential, and friends. Multi-threaded when length(A) >= 100. Will be handled by Yeppp or AppleAccelerate if you load one of them.

ArrayAllez.exp!!

ArrayAllez.exp0

ArrayAllez.exp1

ArrayAllez.exp_ — Function.

exp!(A)
exp_(A) = exp!(similar(A), A)
exp0(A) ≈ exp.(A)

Element-wise in-place exponential, and friends. Multi-threaded when length(A) >= 100. Will be handled by Yeppp or AppleAccelerate if you load one of them.

ArrayAllez.include

ArrayAllez.inv! — Function.

inv!(A) ≈ 1 ./ A
inv!(A, b::Number) ≈ b ./ A

And inv_(A) which copies, and inv0(A) simple broadcasting. Multi-threaded when length(A) >= 1000. Will be handled by AppleAccelerate if you load it.

ArrayAllez.inv!!

ArrayAllez.inv0

ArrayAllez.inv_ — Function.

inv!(A) ≈ 1 ./ A
inv!(A, b::Number) ≈ b ./ A

And inv_(A) which copies, and inv0(A) simple broadcasting. Multi-threaded when length(A) >= 1000. Will be handled by AppleAccelerate if you load it.

ArrayAllez.iscale! — Function.

iscale!(A, b::Number) ≈ A ./ b
iscale!(A, v::Vector) ≈ A ./ v      # A::Matrix
iscale!(A, r::Adjoint) ≈ A ./ r     # r::RowVector / Transpose etc.
iscale!(A, B) ≈ A ./ B

For each of these, there is also iscale_(A, ...) non-mutating but perhaps accellerated, and iscale0(A, ...) simple broadcasting. Finally there is iscale!!(A, x) which mutate both arguments, wihch may be a terrible idea.

ArrayAllez.iscale!!

ArrayAllez.iscale0

ArrayAllez.iscale_ — Function.

iscale!(A, b::Number) ≈ A ./ b
iscale!(A, v::Vector) ≈ A ./ v      # A::Matrix
iscale!(A, r::Adjoint) ≈ A ./ r     # r::RowVector / Transpose etc.
iscale!(A, B) ≈ A ./ B

For each of these, there is also iscale_(A, ...) non-mutating but perhaps accellerated, and iscale0(A, ...) simple broadcasting. Finally there is iscale!!(A, x) which mutate both arguments, wihch may be a terrible idea.

ArrayAllez.load_note

ArrayAllez.log! — Function.

log!(A)
log_(A) ≈ log!(similar(A), A)
log0(A) = log.(A)

Element-wise in-place natural logarithm, and friends. Multi-threaded when length(A) >= 100. Will be handled by Yeppp or AppleAccelerate if you load one of them.

ArrayAllez.log!!

ArrayAllez.log0

ArrayAllez.log1

ArrayAllez.log_ — Function.

log!(A)
log_(A) ≈ log!(similar(A), A)
log0(A) = log.(A)

Element-wise in-place natural logarithm, and friends. Multi-threaded when length(A) >= 100. Will be handled by Yeppp or AppleAccelerate if you load one of them.

ArrayAllez.osizes

ArrayAllez.scale! — Function.

scale!(A, b::Number) ≈ A .* b
scale!(A, v::Vector) ≈ A .* v       # A::Matrix
scale!(A, r::Adjoint) ≈ A .* r      # r::RowVector / Transpose etc.
scale!(A, B) ≈ A .* B

For each of these, there is also also scale_(A, ...) non-mutating but perhaps accellerated, and scale0(A, ...) simple broadcasting.

ArrayAllez.scale!!

ArrayAllez.scale0

ArrayAllez.scale_ — Function.

scale!(A, b::Number) ≈ A .* b
scale!(A, v::Vector) ≈ A .* v       # A::Matrix
scale!(A, r::Adjoint) ≈ A .* r      # r::RowVector / Transpose etc.
scale!(A, B) ≈ A .* B

For each of these, there is also also scale_(A, ...) non-mutating but perhaps accellerated, and scale0(A, ...) simple broadcasting.

ArrayAllez.similar_ — Function.

similar_(name, A)     ≈ similar(A)
Array_{T}(name, size) ≈ Array{T}(undef, size)

New arrays for intermediate results, drawn from an LRU cache when length(A) >= 2000. The cache's key uses name::Symbol as well as type & size to ensure different uses don't collide.

copy_(name, A) = copyto!(similar_(name, A), A)

Just like that.

ArrayAllez.sum_

ArrayAllez.↓

ArrayAllez.⊙