Docstrings

RandomV06.AbstractRNG

RandomV06.Close1Open2

RandomV06.CloseOpen

RandomV06.FloatInterval

RandomV06.GLOBAL_RNG

RandomV06.MTCacheLength

RandomV06.MersenneTwister — Type.

MersenneTwister(seed)

Create a MersenneTwister RNG object. Different RNG objects can have their own seeds, which may be useful for generating different streams of random numbers.

Example

julia> rng = MersenneTwister(1234);

RandomV06.RandomDevice — Type.

RandomDevice()

Create a RandomDevice RNG object. Two such objects will always generate different streams of random numbers.

RandomV06.RandomV06

RandomV06.RangeGenerator

RandomV06.RangeGeneratorBigInt

RandomV06.RangeGeneratorInt

RandomV06.UUID

RandomV06.V06

RandomV06.V07

RandomV06.Vcur

RandomV06.__init__

RandomV06.bitrand — Function.

bitrand([rng=GLOBAL_RNG], [dims...])

Generate a BitArray of random boolean values.

Example

julia> rng = MersenneTwister(1234);

julia> bitrand(rng, 10)
10-element BitArray{1}:
  true
  true
  true
 false
  true
 false
 false
  true
 false
  true

RandomV06.bitrand_ver

RandomV06.dsfmt_gv_srand

RandomV06.eval

RandomV06.fe

RandomV06.fi

RandomV06.fill_array!

RandomV06.gen_rand

RandomV06.globalRNG

RandomV06.include

RandomV06.ke

RandomV06.ki

RandomV06.make_seed

RandomV06.mask128

RandomV06.maxmultiple

RandomV06.maxmultiplemix

RandomV06.mt_avail

RandomV06.mt_empty

RandomV06.mt_pop!

RandomV06.mt_setempty!

RandomV06.mt_setfull!

RandomV06.rand — Function.

rand([rng=GLOBAL_RNG], [S], [dims...])

Pick a random element or array of random elements from the set of values specified by S; S can be

an indexable collection (for example 1:n or ['x','y','z']), or

a type: the set of values to pick from is then equivalent to typemin(S):typemax(S) for integers (this is not applicable to BigInt), and to $[0, 1)$ for floating point numbers;

S defaults to Float64.

RandomV06.rand! — Function.

rand!([rng=GLOBAL_RNG], A, [coll])

Populate the array A with random values. If the indexable collection coll is specified, the values are picked randomly from coll. This is equivalent to copy!(A, rand(rng, coll, size(A))) or copy!(A, rand(rng, eltype(A), size(A))) but without allocating a new array.

Example

julia> rng = MersenneTwister(1234);

julia> rand!(rng, zeros(5))
5-element Array{Float64,1}:
 0.590845
 0.766797
 0.566237
 0.460085
 0.794026

RandomV06.rand_AbstractArray_Float64!

RandomV06.rand_inbounds

RandomV06.rand_lt — Function.

Return a random Int (masked with mask) in $[0, n)$, when n <= 2^52.

RandomV06.rand_ui10_raw

RandomV06.rand_ui23_raw

RandomV06.rand_ui2x52_raw

RandomV06.rand_ui52

RandomV06.rand_ui52_raw

RandomV06.rand_ui52_raw_inbounds

RandomV06.rand_ver

RandomV06.rand_ver!

RandomV06.randcycle — Function.

randcycle([rng=GLOBAL_RNG,] n::Integer)

Construct a random cyclic permutation of length n. The optional rng argument specifies a random number generator, see Random Numbers.

Example

julia> rng = MersenneTwister(1234);

julia> randcycle(rng, 6)
6-element Array{Int64,1}:
 3
 5
 4
 6
 1
 2

RandomV06.randcycle_ver

RandomV06.randexp — Function.

randexp([rng=GLOBAL_RNG], [T=Float64], [dims...])

Generate a random number of type T according to the exponential distribution with scale 1. Optionally generate an array of such random numbers. The Base module currently provides an implementation for the types Float16, Float32, and Float64 (the default).

Examples

julia> rng = MersenneTwister(1234);

julia> randexp(rng, Float32)
2.4835055f0

julia> randexp(rng, 3, 3)
3×3 Array{Float64,2}:
 1.5167    1.30652   0.344435
 0.604436  2.78029   0.418516
 0.695867  0.693292  0.643644

RandomV06.randexp! — Function.

randexp!([rng=GLOBAL_RNG], A::AbstractArray) -> A

Fill the array A with random numbers following the exponential distribution (with scale 1).

Example

julia> rng = MersenneTwister(1234);

julia> randexp!(rng, zeros(5))
5-element Array{Float64,1}:
 2.48351
 1.5167
 0.604436
 0.695867
 1.30652

RandomV06.randexp_unlikely

RandomV06.randexp_ver

RandomV06.randexp_ver!

RandomV06.randn — Function.

randn([rng=GLOBAL_RNG], [T=Float64], [dims...])

Generate a normally-distributed random number of type T with mean 0 and standard deviation 1. Optionally generate an array of normally-distributed random numbers. The Base module currently provides an implementation for the types Float16, Float32, and Float64 (the default).

Examples

julia> rng = MersenneTwister(1234);

julia> randn(rng, Float64)
0.8673472019512456

julia> randn(rng, Float32, (2, 4))
2×4 Array{Float32,2}:
 -0.901744  -0.902914  2.21188   -0.271735
 -0.494479   0.864401  0.532813   0.502334

RandomV06.randn! — Function.

randn!([rng=GLOBAL_RNG], A::AbstractArray) -> A

Fill the array A with normally-distributed (mean 0, standard deviation 1) random numbers. Also see the rand function.

Example

julia> rng = MersenneTwister(1234);

julia> randn!(rng, zeros(5))
5-element Array{Float64,1}:
  0.867347
 -0.901744
 -0.494479
 -0.902914
  0.864401

RandomV06.randn_unlikely

RandomV06.randn_ver

RandomV06.randn_ver!

RandomV06.randperm — Function.

randperm([rng=GLOBAL_RNG,] n::Integer)

Construct a random permutation of length n. The optional rng argument specifies a random number generator (see Random Numbers). To randomly permute a arbitrary vector, see shuffle or shuffle!.

Example

julia> rng = MersenneTwister(1234);

julia> randperm(rng, 4)
4-element Array{Int64,1}:
 2
 1
 4
 3

RandomV06.randperm_ver

RandomV06.randstring

RandomV06.randstring_ver

RandomV06.randsubseq — Function.

randsubseq(A, p) -> Vector

Return a vector consisting of a random subsequence of the given array A, where each element of A is included (in order) with independent probability p. (Complexity is linear in p*length(A), so this function is efficient even if p is small and A is large.) Technically, this process is known as "Bernoulli sampling" of A.

RandomV06.randsubseq!

RandomV06.randsubseq_ver

RandomV06.randsubseq_ver!

RandomV06.rem_knuth

RandomV06.reserve

RandomV06.reserve_1

RandomV06.seed_ver!

RandomV06.shuffle — Function.

shuffle([rng=GLOBAL_RNG,] v)

Return a randomly permuted copy of v. The optional rng argument specifies a random number generator (see Random Numbers). To permute v in-place, see shuffle!. To obtain randomly permuted indices, see randperm.

Example

julia> rng = MersenneTwister(1234);

julia> shuffle(rng, collect(1:10))
10-element Array{Int64,1}:
  6
  1
 10
  2
  3
  9
  5
  7
  4
  8

RandomV06.shuffle! — Function.

shuffle!([rng=GLOBAL_RNG,] v)

In-place version of shuffle: randomly permute the array v in-place, optionally supplying the random-number generator rng.

Example

julia> rng = MersenneTwister(1234);

julia> shuffle!(rng, collect(1:16))
16-element Array{Int64,1}:
  2
 15
  5
 14
  1
  9
 10
  6
 11
  3
 16
  7
  4
 12
  8
 13

RandomV06.shuffle_ver

RandomV06.shuffle_ver!

RandomV06.srand — Function.

srand([rng=GLOBAL_RNG], [seed]) -> rng
srand([rng=GLOBAL_RNG], filename, n=4) -> rng

Reseed the random number generator. If a seed is provided, the RNG will give a reproducible sequence of numbers, otherwise Julia will get entropy from the system. For MersenneTwister, the seed may be a non-negative integer, a vector of UInt32 integers or a filename, in which case the seed is read from a file (4n bytes are read from the file, where n is an optional argument). RandomDevice does not support seeding.

RandomV06.uuid1 — Function.

uuid1([rng::AbstractRNG=GLOBAL_RNG]) -> UUID

Generates a version 1 (time-based) universally unique identifier (UUID), as specified by RFC 4122. Note that the Node ID is randomly generated (does not identify the host) according to section 4.5 of the RFC.

Example

julia> rng = MersenneTwister(1234);

julia> Base.Random.uuid1(rng)
2cc938da-5937-11e7-196e-0f4ef71aa64b

RandomV06.uuid4 — Function.

uuid4([rng::AbstractRNG=GLOBAL_RNG]) -> UUID

Generates a version 4 (random or pseudo-random) universally unique identifier (UUID), as specified by RFC 4122.

Example

julia> rng = MersenneTwister(1234);

julia> Base.Random.uuid4(rng)
82015f10-44cc-4827-996e-0f4ef71aa64b

RandomV06.uuid_version — Function.

uuid_version(u::UUID) -> Integer

Inspects the given UUID and returns its version (see RFC 4122).

Example

julia> rng = MersenneTwister(1234);

julia> Base.Random.uuid_version(Base.Random.uuid4(rng))
4

RandomV06.we

RandomV06.wi

RandomV06.ziggurat_exp_r

RandomV06.ziggurat_nor_inv_r

RandomV06.ziggurat_nor_r