Docstrings

SpectralGaussianSimulation.@pack_SpecGaussSimParam

SpectralGaussianSimulation.@pack_SpecGaussSimParam!

SpectralGaussianSimulation.@unpack_SpecGaussSimParam

SpectralGaussianSimulation.SpecGaussSim — Type.

SpecGaussSim(var₁=>param₁, var₂=>param₂, ...)

Spectral Gaussian simulation (a.k.a. FFT simulation).

Parameters

variogram - theoretical variogram (default to GaussianVariogram())

mean - mean of Gaussian field (default to 0)

Global parameters

threads - number of threads in FFT (default to all physical cores)

References

Gutjahr 1997. General joint conditional simulations using a fast Fourier transform method.

SpectralGaussianSimulation.SpecGaussSimParam — Type.

SpecGaussSim(var₁=>param₁, var₂=>param₂, ...)

Spectral Gaussian simulation (a.k.a. FFT simulation).

Parameters

variogram - theoretical variogram (default to GaussianVariogram())

mean - mean of Gaussian field (default to 0)

Global parameters

threads - number of threads in FFT (default to all physical cores)

References

Gutjahr 1997. General joint conditional simulations using a fast Fourier transform method.

source

SpectralGaussianSimulation.SpectralGaussianSimulation

SpectralGaussianSimulation.eval

SpectralGaussianSimulation.include