Docstrings

ApproxManifoldProducts.*

ApproxManifoldProducts.ApproxManifoldProducts

ApproxManifoldProducts.BallTree

ApproxManifoldProducts.BallTreeDensity

ApproxManifoldProducts.CTs

ApproxManifoldProducts.Circular

ApproxManifoldProducts.Euclid2

ApproxManifoldProducts.EuclideanManifold

ApproxManifoldProducts.KDE

ApproxManifoldProducts.KernelDensityEstimate

ApproxManifoldProducts.MB

ApproxManifoldProducts.Manifold

ApproxManifoldProducts.ManifoldBelief

ApproxManifoldProducts.MixtureDensity

ApproxManifoldProducts.Ndim

ApproxManifoldProducts.Npts

ApproxManifoldProducts.TUs

ApproxManifoldProducts.VectorRange

ApproxManifoldProducts.__init__

ApproxManifoldProducts.addtheta

ApproxManifoldProducts.convert

ApproxManifoldProducts.difftheta

ApproxManifoldProducts.eval

ApproxManifoldProducts.evalAvgLogL

ApproxManifoldProducts.evaluateDualTree

ApproxManifoldProducts.evaluateManifoldNaive1D! — Function.

evaluateManifoldNaive1D!(ret, idx, pts, bw, x)
evaluateManifoldNaive1D!(ret, idx, pts, bw, x, loo)
evaluateManifoldNaive1D!(ret, idx, pts, bw, x, loo, diffop)

Evalute the KDE naively as equally weighted Gaussian kernels with common bandwidth. This function does, however, allow on-manifold evaluations.

ApproxManifoldProducts.get2DLambda

ApproxManifoldProducts.get2DMu

ApproxManifoldProducts.get2DMuMin

ApproxManifoldProducts.getBW

ApproxManifoldProducts.getCircLambda

ApproxManifoldProducts.getCircMu

ApproxManifoldProducts.getKDEMax

ApproxManifoldProducts.getKDEMean

ApproxManifoldProducts.getKDERange

ApproxManifoldProducts.getKDEfit

ApproxManifoldProducts.getPoints

ApproxManifoldProducts.getWeights

ApproxManifoldProducts.include

ApproxManifoldProducts.intersIntgAppxIS

ApproxManifoldProducts.kde!

ApproxManifoldProducts.kde!_CircularNaiveCV

ApproxManifoldProducts.kld

ApproxManifoldProducts.logmap_SO2

ApproxManifoldProducts.manifoldLooCrossValidation — Function.

manifoldLooCrossValidation(pts, bw; own, diffop)

Calculate negative entropy with leave one out (j'th element) cross validation.

Background

From: Silverman, B.: Density Estimation for Statistics and Data Analysis, 1986, p.52

Probability density function p(x), as estimated by kernels

\[hatp_{-j}(x) = 1/(N-1) Σ_{i != j}^N frac{1}{sqrt{2pi}σ } exp{ -frac{(x-μ)^2}{2 σ^2} }\]

and has Cross Validation number as the average log evaluations of leave one out hatp_{-j}(x):

\[CV(p) = 1/N Σ_i^N log hat{p}_{-j}(x_i)\]

This quantity CV is related to an entropy H(p) estimate via:

\[H(p) = -CV(p)\]

ApproxManifoldProducts.marginal

ApproxManifoldProducts.normDistAccAt! — Function.

normDistAccAt!(ret, idx, x, sigma)
normDistAccAt!(ret, idx, x, sigma, w)

Probability density function p(x), as estimated by kernels

\[hatp_{-j}(x) = 1/(N-1) Σ_{i != j}^N frac{1}{sqrt{2pi}σ } exp{ -frac{(x-μ)^2}{2 σ^2} }\]

ApproxManifoldProducts.prodAppxMSGibbsS

ApproxManifoldProducts.rand

ApproxManifoldProducts.rbf

ApproxManifoldProducts.rbf!

ApproxManifoldProducts.rbfAccAt!

ApproxManifoldProducts.reci_s2pi

ApproxManifoldProducts.resample

ApproxManifoldProducts.resid2DLinear

ApproxManifoldProducts.root

ApproxManifoldProducts.sample

ApproxManifoldProducts.setForceEvalDirect!

ApproxManifoldProducts.solveresid2DLinear

ApproxManifoldProducts.solveresid2DLinear!

ApproxManifoldProducts.string