Docstrings

InteractiveChaos.@SMatrix

InteractiveChaos.@SVector

InteractiveChaos.@windowed

InteractiveChaos.AbstractDataset

InteractiveChaos.AbstractEmbedding

InteractiveChaos.AbstractNeighborhood

InteractiveChaos.AbstractRecurrenceMatrix

InteractiveChaos.CDS_KWARGS

InteractiveChaos.ChaosTools

InteractiveChaos.Chebyshev

InteractiveChaos.Cityblock

InteractiveChaos.ContinuousDynamicalSystem

InteractiveChaos.CrossRecurrenceMatrix

InteractiveChaos.Dataset

InteractiveChaos.DelayEmbedding

InteractiveChaos.DelayEmbeddings

InteractiveChaos.DiscreteDynamicalSystem

InteractiveChaos.DynamicalSystem

InteractiveChaos.DynamicalSystems

InteractiveChaos.DynamicalSystemsBase

InteractiveChaos.Euclidean

InteractiveChaos.FixedMassNeighborhood

InteractiveChaos.FixedSizeNeighborhood

InteractiveChaos.InteractiveChaos

InteractiveChaos.JointRecurrenceMatrix

InteractiveChaos.KDTree

InteractiveChaos.MTDelayEmbedding

InteractiveChaos.PlaneCrossing

InteractiveChaos.RecurrenceAnalysis

InteractiveChaos.RecurrenceMatrix

InteractiveChaos.SMatrix

InteractiveChaos.SVector

InteractiveChaos.Size

InteractiveChaos.Systems

InteractiveChaos._randomcolor

InteractiveChaos.autocor

InteractiveChaos.boxcounting_dim

InteractiveChaos.broomhead_king

InteractiveChaos.capacity_dim

InteractiveChaos.columns

InteractiveChaos.controlwindow — Function.

controlwindow(D, n, Ttr, density, i)

Create an Electron control window for the orbit diagram interactive application.

return n, Ttr, density, i, ▢update, ▢back, ▢reset, α,
       ⬜pmin, ⬜pmax, ⬜umin, ⬜umax

All returned values are Observables. Their value corresponds to the one chosen in the Electron window. Items with ▢ are buttons and with ⬜ are the boxes with limits.

InteractiveChaos.coordinates

InteractiveChaos.determinism

InteractiveChaos.dimension

InteractiveChaos.distancematrix

InteractiveChaos.divergence

InteractiveChaos.dl_average

InteractiveChaos.dl_entropy

InteractiveChaos.dl_max

InteractiveChaos.embed

InteractiveChaos.estimate_boxsizes

InteractiveChaos.estimate_delay

InteractiveChaos.estimate_dimension

InteractiveChaos.eval

InteractiveChaos.exponential_decay_fit

InteractiveChaos.gali

InteractiveChaos.genentropy

InteractiveChaos.generalized_dim

InteractiveChaos.get_deviations

InteractiveChaos.get_state

InteractiveChaos.grayscale

InteractiveChaos.include

InteractiveChaos.information_dim

InteractiveChaos.integrator

InteractiveChaos.interactive_orbitdiagram — Function.

interactive_orbitdiagram(ds::DiscreteDynamicalSystem,
    i::Int, p_index, p_min, p_max;
    density = 500, u0 = get_state(ds), Ttr = 200, n = 500,
    parname = "p"
)

Open an interactive application for exploring orbit diagrams (ODs) of discrete systems. The functionality works for any discrete system.

Once initialized it opens a Makie plot window and an Electron control window.

Interaction

By using the Electron window you are able to update all parameters of the OD interactively (like e.g. n or Ttr). You have to press update after changing these parameters. You you can even decide which variable to get the OD for, by choosing one of the variables from the wheel (again, press update afterwards).

In the Makie window you can interactively zoom into the OD. Click and drag with the left mouse button to select a region in the OD. This region is then re-computed at a higher resolution (i.e. we don't "just zoom").

Back in the Electron window, you can press reset to bring the OD in the original state (and variable). Pressing back will go back through the history of your exploration History is stored when any change happens (besides transparency).

Accessing the data

What is plotted on the application window is a true orbit diagram, not a plotting shorthand. This means that all data are obtainable and usable directly. Internally we always scale the orbit diagram to [0,1]² (to allow Float64 precision even though plotting is Float32-based). This however means that it is necessary to transform the data in real scale. This is done through the function scaleod which accepts the 5 arguments returned from the current function:

od, pmin, pmax, umin, umax = interactive_orbitdiagram(...)
ps, us = scaleod(od, pmin, pmax, umin, umax)

InteractiveChaos.interactivepsos — Function.

interactivepsos(cds, plane, idxs, complete; kwargs...)

Open an interactive application for exploring a Poincaré surface of section (PSOS) of the continuous dynamical system cds.

The plane can only be the Tuple type accepted by poincaresos, i.e. (i, r) for the ith variable crossing the value r. idxs gives the two indices of the variables to be displayed, since the PSOS plot is always a 2D scatterplot. I.e. idxs = (1, 2) will plot the 1st versus second variable of the PSOS. It follows that plane[1] ∉ idxs must be true.

complete is a three-argument function that completes the new initial state during interactive use, see below.

Keyword Arguments

direction, Ttr, rootkw : Same use as in poincaresos.

tfinal : A range of values for the total integration time (chosen interactively).

color : A function of the system's initial condition, that returns a color to plot the new points with. A random color is chosen by default.

makiekwargs : A NamedTuple of keyword arguments passed to AbstractPlotting.scatter.

diffeq... : Any extra keyword arguments are passed into init of DiffEq.

Interaction

The application is a standard AbstractPlotting scatterplot, which shows the PSOS of the system, initially using the system's u0. Two sliders control the final evolution time and the size of the marker points.

Upon clicking within the bounds of the scatter plot your click is transformed into a new initial condition, which is further evolved and its PSOS is computed and then plotted into the scatter plot.

Your click is transformed into a full D-dimensional initial condition through the function complete. The first two arguments of the function are the positions of the click on the PSOS. The third argument is the value of the variable the PSOS is defined on. To be more exact, this is how the function is called:

x, y = mouseclick; z = plane[2]
newstate = complete(x, y, z)

The complete function can throw an error for ill-condition x, y, z. This will be properly handled instead of breaking the application. This newstate is also given to the function color that gets a new color for the new points.

InteractiveChaos.jacobian

InteractiveChaos.kaplanyorke_dim

InteractiveChaos.lambdamatrix

InteractiveChaos.lambdaperms

InteractiveChaos.laminarity

InteractiveChaos.linear_region

InteractiveChaos.linear_regions

InteractiveChaos.linreg

InteractiveChaos.lyapunov

InteractiveChaos.lyapunovs

InteractiveChaos.maxima

InteractiveChaos.meanrecurrencetime

InteractiveChaos.min_pairwise_distance

InteractiveChaos.minima

InteractiveChaos.minimal_normaized_od — Function.

minimal_normaized_od(integ, i, p_index, pmin, pmax,
                     density, n, Ttr, u0)
minimal_normaized_od(integ, i, p_index, pmin, pmax,
                     density, n, Ttr, u0, xmin, xmax)

Compute and return a minimal and normalized orbit diagram (OD).

All points are stored in a single vector of Point2f0 to ensure fastest possible plotting. In addition all numbers are scaled to [0, 1]. This allows us to have 64-bit precision while display is only 32-bit!

The version with xmin, xmax only keeps points with limits between the real xmin, xmax (in the normal units of the dynamical system).

InteractiveChaos.minmaxima

InteractiveChaos.mutualinformation

InteractiveChaos.neighborhood

InteractiveChaos.nmprt

InteractiveChaos.non0hist

InteractiveChaos.numericallyapunov

InteractiveChaos.orbitdiagram

InteractiveChaos.orthonormal

InteractiveChaos.parallel_integrator

InteractiveChaos.periodicorbits

InteractiveChaos.permentropy

InteractiveChaos.poincaresos

InteractiveChaos.produce_orbitdiagram

InteractiveChaos.reconstruct

InteractiveChaos.recurrenceplot

InteractiveChaos.recurrencerate

InteractiveChaos.recurrencestructures

InteractiveChaos.reinit!

InteractiveChaos.rqa

InteractiveChaos.rt_average

InteractiveChaos.rt_entropy

InteractiveChaos.rt_max

InteractiveChaos.scaleod — Function.

scaleod(od, pmin, pmax, umin, umax) -> ps, us

Given the return values of interactive_orbitdiagram, produce orbit diagram data scaled correctly in data units. Return the data as a vector of parameter values and a vector of corresponding variable values.

InteractiveChaos.set_deviations!

InteractiveChaos.set_parameter!

InteractiveChaos.set_state!

InteractiveChaos.sorteddistances

InteractiveChaos.step!

InteractiveChaos.stochastic_indicator

InteractiveChaos.subscript — Function.

subscript(i::Int)

Transform i to a string that has i as a subscript.

InteractiveChaos.tangent_integrator

InteractiveChaos.textrecurrenceplot

InteractiveChaos.trajectory

InteractiveChaos.trappingtime

InteractiveChaos.trend

InteractiveChaos.update_controls!

InteractiveChaos.vl_average

InteractiveChaos.vl_entropy

InteractiveChaos.vl_max