Docstrings

TaylorModels...

TaylorModels.@I_str

TaylorModels.@biginterval

TaylorModels.@decorated

TaylorModels.@floatinterval

TaylorModels.@format

TaylorModels.@interval

TaylorModels.@make_interval

TaylorModels.AbstractInterval

TaylorModels.AbstractSeries

TaylorModels.DecoratedInterval

TaylorModels.HomogeneousPolynomial

TaylorModels.Interval

TaylorModels.IntervalArithmetic

TaylorModels.IntervalBox

TaylorModels.NumberNotSeries

TaylorModels.RTaylorModel1 — Type.

RTaylorModel1{T,S}

Taylor model in 1 variable, providing a rigurous polynomial approximation (around x_0) and a relative remainder \delta for a function f(x) in one variable, valid in the interval I. Corresponds to definition 2.3.2 of Mioara Joldes' thesis.

Fields:

pol: polynomial approximation, represented as TaylorSeries.Taylor1

rem: the interval bound

x0: expansion point

I: interval over which the Taylor model is defined / valid

The approximation f(x) = \sum_i p_i (x - x_0)^i + \delta (x - x_0)^{n+1} is satisfied for all x\in I; n is the order (degree) of the polynomial p(x).

TaylorModels.Region

TaylorModels.RoundTiesToAway

TaylorModels.RoundTiesToEven

TaylorModels.Taylor1

TaylorModels.TaylorModel1 — Type.

TaylorModel1{T,S}

Taylor model in 1 variable, providing a rigurous polynomial approximation (around x_0) and an absolute remainder \Delta for a function f(x) in one variable, valid in the interval I. Corresponds to definition 2.1.3 of Mioara Joldes' thesis.

Fields:

pol: polynomial approximation, represented as TaylorSeries.Taylor1

rem: the interval bound

x0: expansion point

I: interval over which the Taylor model is defined / valid

The approximation f(x) = \sum_{i=0}^n p_i (x - x_0)^i + \Delta is satisfied for all x\in I (0\in\Delta); n is the order (degree) of the polynomial p(x).

TaylorModels.TaylorModelN — Type.

TaylorModelN{N,T,S}

Taylor Models with absolute remainder for N independent variables.

TaylorModels.TaylorModels

TaylorModels.TaylorN

TaylorModels.TaylorSeries

TaylorModels._evaluate

TaylorModels._rpaar — Function.

_rpaar(f::Function, x0::Interval, I::Interval, _order::Integer)

Rigurous polynomial approximation (RPA) with absolute remainder for the function f on the interval I, using a Taylor expansion around the interval x0 of order _order. The bound is computed by bound_absrem(@ref) exploiting monotonicity if possible, otherwise, it uses Lagrange bound.

TaylorModels._rparr — Function.

_rparr(f::Function, x0::Interval, I::Interval, _order::Integer)

Rigurous polynomial approximation (RPA) with relative remainder for the function f on the interval I, using a Taylor expansion around the interval x0 of order _order. The bound is computed by bound_relrem(@ref) exploiting monotonicity if possible, otherwise, it uses the Lagrange coefficient.

TaylorModels.basediv

TaylorModels.bisect

TaylorModels.bound_absrem — Function.

bound_absrem(f::Function, polf::Taylor1, polfI::Taylor1, x0::Interval, I::Interval)

Bound the absolute remainder of the polynomial approximation of f given by the Taylor polynomial polf around x0 on the interval I. It requires the interval extension polfI of the polynomial that approximates f for the whole interval I, in order to compute the Lagrange remainder.

If polfI[end] has a definite sign, then it is monotonic in the intervals [I.lo, x0] and [x0.hi, I.hi], which is exploited; otherwise, it is used to compute the Lagrange remainder.

TaylorModels.bound_relrem — Function.

bound_relrem(f::Function, polf::Taylor1, polfI::Taylor1, x0::Interval, I::Interval)

Bound the relative remainder of the polynomial approximation of f given by the Taylor polynomial polf around x0 on the interval I. It requires an the interval extension polfI of a polynomial that approximates f for the whole interval I, in order to compute the Lagrange remainder.

If polfI[end] has a definite sign, then it is monotonic in the interval I, which is exploited; otherwise, the last coefficients bounds the relative remainder.

TaylorModels.bound_taylor1 — Function.

bound_taylor1(fT::Taylor1, I::Interval)

Compute a tight polynomial bound for the Taylor polynomial fT in the interval I.

Note: Algorithm 2.1.1 corresponds to evaluate(fT, I) or simply fT(I). This function uses the roots of the derivative offt` to obtain a tighter bound.

bound_taylor1(fT::Taylor1, fTd::Taylor1, I::Interval)

Compute a tight polynomial bound for the Taylor polynomial fT in the interval I, considering whether its derivative ftd has a definite sign.

TaylorModels.cancelminus

TaylorModels.cancelplus

TaylorModels.check_existence — Function.

check_existence(f, tm::T, xm::T, x0::Interval, x_test::Interval)

Checks that the range of one iterate of the Picard-Lindelöf operator is contained in the a-priori interval x_test (of the dependent variable) that bounds the solution of the ODE defined by f. This function returns an interval of the independent variable where the a-priori solution is warranted to exist; see shrink_for_existance. Here, tm and xm are Taylor Models (TaylorModel1 or RTaylorModel1) defined for the independent and dependent variables, and x0 is the initial condition.

TaylorModels.com

TaylorModels.constant_term

TaylorModels.contains_zero

TaylorModels.dac

TaylorModels.decoration

TaylorModels.def

TaylorModels.derivative

TaylorModels.diam

TaylorModels.differentiate

TaylorModels.displayBigO

TaylorModels.displaymode

TaylorModels.dist

TaylorModels.dot

TaylorModels.emptyinterval

TaylorModels.entireinterval

TaylorModels.eps

TaylorModels.eval

TaylorModels.evaluate

TaylorModels.evaluate!

TaylorModels.extended_div

TaylorModels.fnlist

TaylorModels.fp_rpa — Function.

fp_rpa(tm::TaylorModel1{Interval{T},T})
fp_rpa(tm::RTaylorModel1{Interval{T},T})

Convert a tm TaylorModel to a TaylorModel whose polynomial coefficients of type T<:Real. The accumulated error is added to the remainder. The mid point of the expansion interval is preferentially rounded down if it is not an exactly representable value.

TaylorModels.get_numvars

TaylorModels.get_order

TaylorModels.get_variable_names

TaylorModels.get_variable_symbols

TaylorModels.get_variables

TaylorModels.getcoeff

TaylorModels.hull

TaylorModels.ill

TaylorModels.include

TaylorModels.inf

TaylorModels.infimum

TaylorSeries.integrate — Function.

integrate(a::T, c0::Interval)

Integrates the one-variable Taylor Model (TaylorModel1 or RTaylorModel1) with respect to the independent variable; c0 is the interval representing the integration constant; if omitted it is considered as the zero interval.

TaylorModels.interval

TaylorModels.interval_from_midpoint_radius

TaylorModels.interval_part

TaylorModels.inverse

TaylorModels.isatomic

TaylorModels.iscommon

TaylorModels.isdisjoint

TaylorModels.isempty

TaylorModels.isentire

TaylorModels.isinterior

TaylorModels.isnai

TaylorModels.isthin

TaylorModels.isunbounded

TaylorModels.mag

TaylorModels.mid

TaylorModels.midpoint_radius

TaylorModels.mig

TaylorModels.nai

TaylorModels.parameters

TaylorModels.pi_interval

TaylorModels.picard_lindelöf — Function.

picard_lindelöf(f, tm::T, xm::T, x0::Interval)
𝒫(f, tm::T, xm::T, x0::Interval)

Returns the application of the Picard-Lindelöf operator associated to the ODE $\dot{x} = f(t,x)$, with initial condition x0. Here, tm and xm are (one-variable) Taylor Models (TaylorModel1 or RTaylorModel1).

𝒫 is an abbreviation of this operator, which is obtained as \mscrP<TAB>.)

TaylorModels.polynomial

TaylorModels.pow

TaylorModels.precedes

TaylorModels.radius

TaylorModels.reducetoorder — Function.

reducetoorder(a::RTaylorModel1, m::Integer)

From a::RTaylorModel1, it returns a RTaylorModel1 of order m.

TaylorModels.remainder

TaylorModels.rpa — Function.

rpa(g::Function, tmf::TaylorModel1) rpa(g::Function, tmf::TaylorModelN)

Rigurous polynomial approximation (RPA) for the function g using the Taylor Model with absolute remainder tmf. The bound is computed exploiting monotonicity if possible, otherwise, it uses Lagrange bound.

rpa(g::Function, tmf::RTaylorModel1)

Rigurous polynomial approximation (RPA) for the function g using the Taylor Model with absolute remainder tmf. The bound is computed exploiting monotonicity if possible, otherwise, it uses Lagrange bound.

TaylorModels.set_variables

TaylorModels.setformat

TaylorModels.setindex

TaylorModels.show_monomials

TaylorModels.show_params_TaylorN

TaylorModels.showfull

TaylorModels.shrink_for_existance — Function.

shrink_for_existance(xm::T, t_interval, x_test, max_steps::Integer=20)

Shrinks the a-priori independent-variable interval t_interval so the range of xm, the Taylor Model (TaylorModel1 or RTaylorModel1) associated with the dependent variable, is contained in the a-priori interval x_test. The method used is some sort of bisection. If no independent-variable interval is found within max_steps, an empty interval is returned.

TaylorModels.strictprecedes

TaylorModels.sup

TaylorModels.supremum

TaylorModels.taylor_expand

TaylorModels.tight_remainder — Function.

tight_remainder(f, tm::T, xm::T, x0::Interval, max_steps::Integer=20)

Returns a Taylor Model for the dependent variable, with a tighter remainder, which is obtained after successive iteration of the Picard-Lindelöf. If the remainder is not tighter (and identity with the former iterate is not obtained) a Taylor Model with an empty interval is returned.

TaylorModels.tmdata

TaylorModels.trv

TaylorModels.tupleTMs

TaylorModels.update!

TaylorModels.use_show_default

TaylorModels.widen

TaylorModels.±

TaylorModels.×

TaylorModels.α_mid

TaylorModels.∅

TaylorModels.∇

TaylorModels.∞

TaylorModels.≺

TaylorModels.⊂

TaylorModels.⊃

TaylorModels.⊇

TaylorModels.⪽

TaylorModels.𝒫 — Function.

picard_lindelöf(f, tm::T, xm::T, x0::Interval)
𝒫(f, tm::T, xm::T, x0::Interval)

Returns the application of the Picard-Lindelöf operator associated to the ODE $\dot{x} = f(t,x)$, with initial condition x0. Here, tm and xm are (one-variable) Taylor Models (TaylorModel1 or RTaylorModel1).

𝒫 is an abbreviation of this operator, which is obtained as \mscrP<TAB>.)