Docstrings

SimpleGF2.GF2

SimpleGF2.SimpleGF2

SimpleGF2.add_row_to_row! — Function.

add_row_to_row!(A,i,j) adds row i to row j in the matrix A.

SimpleGF2.eval

SimpleGF2.include

SimpleGF2.nullity — Function.

nullity(A) returns the dimension of the nullspace of A.

SimpleGF2.rref — Function.

rref(A) returns the row reduced echelon form of A.

SimpleGF2.rref! — Function.

rref!(A) overwrites A with its row reduced echelon form.

SimpleGF2.solve — Function.

solve(A,b) returns a solution x to the linear system A*x == b or throws an error if no solution can be found.

SimpleGF2.solve_all — Function.

solve_all(A,b) returns a solution to A*x==b together with a basis for the nullspace of A.

SimpleGF2.solve_augmented

SimpleGF2.swap_rows! — Function.

swap_rows!(A,i,j) swaps rows i and j in the matrix A.