Docstrings

UnicodeGraphics.UnicodeGraphics — Module.

Block and braile rendering of julia arrays, for terminal graphics.

UnicodeGraphics.block_array!

UnicodeGraphics.blockize — Function.

brailize(a, cutoff=0)

Convert an array to a block unicode string, filling values above the cutoff point.

UnicodeGraphics.blockize! — Function.

blockize!(out, a, cutoff=0)

Convert an array to a braile unicode string, filling the out array. Calculation of array dims is a little complicated:

yrange, xrange = axes(a)
out = Array{Char,2}(undef, length(xrange) + 1, (length(yrange) - 1) ÷ 2 + 1)

UnicodeGraphics.braile_array!

UnicodeGraphics.braile_hex

UnicodeGraphics.brailize — Function.

brailize(a, cutoff=0)

Convert an array to a braile unicode string, filling values above the cutoff point.

UnicodeGraphics.brailize! — Function.

brailize!(out, a, cutoff=0)

Convert an array to a braile unicode string, filling the out array. Calculation of array dims is a little complicated:

yrange, xrange = axes(a)
out = Array{Char,2}(undef, (length(xrange) - 1) ÷ 2 + 2, (length(yrange) - 1) ÷ 4 + 1)

UnicodeGraphics.checkval

UnicodeGraphics.eval

UnicodeGraphics.include