Docstrings

ComplexPhasePortrait...

ComplexPhasePortrait.ComplexPhasePortrait

ComplexPhasePortrait.PTcgrid

ComplexPhasePortrait.PTproper

ComplexPhasePortrait.PTstepmod

ComplexPhasePortrait.PTstepphase

ComplexPhasePortrait.Phase

ComplexPhasePortrait.PhasePlot

ComplexPhasePortrait.PortraitType

ComplexPhasePortrait._range

ComplexPhasePortrait.baseArgs

ComplexPhasePortrait.baseColorMap — Function.

Color map for complex phase portrait with 600 elements.

ComplexPhasePortrait.brighten

ComplexPhasePortrait.eval

ComplexPhasePortrait.include

ComplexPhasePortrait.phase — Function.

ComplexPhasePortrait.phase(args...; attributes...)

Command works on plot args 1:N and accepts keyword arguments to style the plot. Creates a new scene!

ComplexPhasePortrait.phase(attributes::Attributes, args...; attributes...)

Like ComplexPhasePortrait.phase(args...; attributes...) but accepts a theme as first argument. Creates a new scene!

ComplexPhasePortrait.phase! — Function.

ComplexPhasePortrait.phase!(args...; attributes...)

Command works on plot args 1:N and accepts keyword arguments to style the plot. Adds new plot to current_scene()

ComplexPhasePortrait.phase!(scene::SceneLike, args...; attributes...)

Command works on plot args 1:N and accepts keyword arguments to style the plot. Adds new plot to scene!

ComplexPhasePortrait.phase!(attributes::Attributes, args...; attributes...)

Like ComplexPhasePortrait.phase!(args...; attributes...) but accepts a theme as first argument. Adds new plot to current_scene()!

ComplexPhasePortrait.phase!(attributes::Attributes, args...; attributes...)

Like ComplexPhasePortrait.phase!(scene, args...; attributes...) but accepts a theme as second argument. Adds new plot to scene!

ComplexPhasePortrait.phaseToImage!

ComplexPhasePortrait.phaseplot

ComplexPhasePortrait.phaseplot!

ComplexPhasePortrait.portrait

ComplexPhasePortrait.sawfun — Function.

Sawtooth function over reals with period dx onto [a,b].

ComplexPhasePortrait.setupPhase

ComplexPhasePortrait.stepfun — Function.

Integer step function with period 1 such that [0,1]⟶[1,nmax].