Docstrings

StratiGraphics.@pack_StratSimParam

StratiGraphics.@pack_StratSimParam!

StratiGraphics.@unpack_StratSimParam

StratiGraphics.Environment — Type.

Environment(processes, transitions, durations)

Geological environment with processes, transitions and durations.

StratiGraphics.ExponentialDuration

StratiGraphics.LandState — Type.

LandState(land)

A geological state consisting of a landscape map.

StratiGraphics.Process

StratiGraphics.Record — Type.

Record(states)

A geological record of geological states.

StratiGraphics.State — Type.

State

A geological state to be evolved by processes.

StratiGraphics.StratSim — Type.

StratSim(var₁=>param₁, var₂=>param₂, ...)

Stratigraphy simulation with Markov-Poisson sampling.

Parameters

environment - geological environment

state - initial geological state

stack - stacking scheme (:erosional or :depositional)

nepochs - number of epochs (default to 10)

References

Hoffimann 2018. Morphodynamic analysis and statistical synthesis of geormorphic data.

StratiGraphics.StratSimParam — Type.

StratSim(var₁=>param₁, var₂=>param₂, ...)

Stratigraphy simulation with Markov-Poisson sampling.

Parameters

environment - geological environment

state - initial geological state

stack - stacking scheme (:erosional or :depositional)

nepochs - number of epochs (default to 10)

References

Hoffimann 2018. Morphodynamic analysis and statistical synthesis of geormorphic data.

source

StratiGraphics.Strata — Type.

Strata(record)

Stratigraphic model from geological record.

StratiGraphics.StratiGraphics

StratiGraphics.UniformDuration

StratiGraphics.eval

StratiGraphics.evolve! — Function.

evolve!(state, proc, Δt)

Evolve the state with process proc for a time period Δt.

StratiGraphics.include

StratiGraphics.simulate — Function.

simulate(env, state, nepochs=10)

Simulate the geological environment env for a number of epochs nepochs starting from a state.

StratiGraphics.voxelize