Library

Library

Constructing a Fun

ApproxFun.Fun — Type.

GaussWeight(Hermite(L), L) is a space spanned by exp(-Lx²) * H_k(sqrt(L) * x) where H_k(x)'s are Hermite polynomials.

GaussWeight() is equivalent to GaussWeight(Hermite(), 1.0) by default.

Fun(s::Space,coefficients::AbstractVector)

returns a Fun with the specified coefficients in the space s

Fun(f,s::Space)

return a Fun representing the function, number, or vector f in the space s. If f is vector-valued, it returns a vector-valued analogue of s.

Fun(f,d::Domain)

returns Fun(f,Space(d)), that is, it uses the default space for the specified domain.

Fun(s::Space)

returns Fun(identity,s)

Fun(f)

returns Fun(f,Chebyshev())

Fun()

returns Fun(identity,Chebyshev()).

Base.ones — Method.

ones(d::Space)

Return the Fun that represents the function one on the specified space.

Base.zeros — Method.

zeros(d::Space)

Return the Fun that represents the function one on the specified space.

Domains

ApproxFun.Arc — Type.

Arc(c,r,(θ₁,θ₂))

represents the arc centred at c with radius r from angle θ₁ to θ₂.

ApproxFun.Circle — Type.

Circle(c,r,o)

represents the circle centred at c with radius r which is positively (o=true) or negatively (o=false) oriented.

ApproxFun.Curve — Constant.

Curve Represents a domain defined by the image of a Fun. Example usage would be

x=Fun(1..2)
Curve(exp(im*x))  # represents an arc

ApproxFun.Disk — Type.

Disk(c,r)

represents the disk centred at c with radius r.

ApproxFun.Segment — Type.

Segment(a,b)

represents a line segment from a to b. In the case where a and b are real and a < b, then this is is equivalent to an Interval(a,b).

Interval

ApproxFun.Line — Type.

Line{a}(c)

represents the line at angle a in the complex plane, centred at c.

ApproxFun.PeriodicSegment — Type.

PeriodicSegment(a,b)

represents a periodic interval from a to b, that is, the point b is identified with a.

ApproxFun.Point

ProductDomain

ApproxFun.Ray — Type.

Ray{a}(c,o)

represents a ray at angle a starting at c, with orientation out to infinity (o = true) or back from infinity (o = false).

UnionDomain

∂

Accessing information about a spaces

ApproxFun.canonicalspace — Function.

canonicalspace(s::Space)

returns a space that is used as a default to implement missing functionality, e.g., evaluation. Implement a Conversion operator or override coefficients to support this.

ApproxFun.itransform — Function.

itransform(s::Space,coefficients::AbstractVector)

Transform coefficients back to values. Defaults to using canonicalspace as in transform.

ApproxFun.transform — Function.

transform(s::Space,vals::Vector)

Transform values on the grid specified by points(s,length(vals)) to coefficients in the space s. Defaults to coefficients(transform(canonicalspace(space),values),canonicalspace(space),space)

DomainSets.dimension — Method.

dimension(s::Space)

returns the dimension of s, which is the maximum number of coefficients.

Inbuilt spaces

ApproxFun.SequenceSpace — Type.

SequenceSpace is the space of all sequences, i.e., infinite vectors. Also denoted ℓ⁰.

ApproxFun.ConstantSpace — Type.

ConstantSpace is the 1-dimensional scalar space.

ApproxFun.Chebyshev — Type.

Chebyshev() is the space spanned by the Chebyshev polynomials

    T_0(x),T_1(x),T_2(x),…

where T_k(x) = cos(k*acos(x)). This is the default space as there exists a fast transform and general smooth functions on [-1,1] can be easily resolved.

ApproxFun.Jacobi — Type.

Jacobi(b,a) represents the space spanned by Jacobi polynomials P_k^{(a,b)}, which are orthogonal with respect to the weight (1+x)^β*(1-x)^α

ApproxFun.Ultraspherical — Type.

Ultraspherical(λ) is the space spanned by the ultraspherical polynomials

    C_0^{(λ)}(x),C_1^{(λ)}(x),C_2^{(λ)}(x),…

Note that λ=1 this reduces to Chebyshev polynomials of the second kind: C_k^{(1)}(x) = U_k(x).

ApproxFun.Taylor — Type.

Taylor() is the space spanned by [1,z,z^2,...]. This is a type alias for Hardy{true}.

ApproxFun.Hardy — Type.

Hardy{false}() is the space spanned by [1/z,1/z^2,...]. Hardy{true}() is the space spanned by [1,z,z^2,...].

ApproxFun.Fourier — Type.

Fourier() is the space spanned by the trigonemtric polynomials

    1,sin(θ),cos(θ),sin(2θ),cos(2θ),…

See also Laurent.

ApproxFun.Laurent — Type.

Laurent() is the space spanned by the complex exponentials

    1,exp(-im*θ),exp(im*θ),exp(-2im*θ),…

See also Fourier.

ApproxFun.CosSpace — Type.

CosSpace() is the space spanned by [1,cos θ,cos 2θ,...]

ApproxFun.SinSpace — Type.

SinSpace() is the space spanned by [sin θ,sin 2θ,...]

ApproxFun.JacobiWeight — Type.

JacobiWeight(β,α,s::Space)

weights a space s by a Jacobi weight, which on -1..1 is (1+x)^β*(1-x)^α. For other domains, the weight is inferred by mapping to -1..1.

ApproxFun.LogWeight — Type.

LogWeight(β,α,s::Space)

represents a function on -1..1 weighted by log((1+x)^β*(1-x)^α). For other domains, the weight is inferred by mapping to -1..1.

ApproxFun.ArraySpace — Type.

ArraySpace(s::Space,dims...)

is used to represent array-valued expansions in a space s. The coefficients are of each entry are interlaced.

For example,

f = Fun(x->[exp(x),sin(x)],-1..1)
space(f) == ArraySpace(Chebyshev(),2)

ApproxFun.TensorSpace — Type.

TensorSpace(a::Space,b::Space)

represents a tensor product of two 1D spaces a and b. The coefficients are interlaced in lexigraphical order.

For example, consider

Fourier()*Chebyshev()  # returns TensorSpace(Fourier(),Chebyshev())

This represents functions on [-π,π) x [-1,1], using the Fourier basis for the first argument and Chebyshev basis for the second argument, that is, φ_k(x)T_j(y), where

φ_0(x) = 1,
φ_1(x) = sin x,
φ_2(x) = cos x,
φ_3(x) = sin 2x,
φ_4(x) = cos 2x
…

By Choosing (k,j) appropriately, we obtain a single basis:

φ_0(x)T_0(y) (= 1),
φ_0(x)T_1(y) (= y),
φ_1(x)T_0(y) (= sin x),
φ_0(x)T_2(y), …

Accessing information about a Fun

ApproxFun.domain — Function.

domain(f::Fun)

returns the domain that f is defined on

ApproxFun.coefficients — Function.

coefficients(f::Fun) -> Vector

returns the coefficients of f, corresponding to the space space(f).

coefficients(f::Fun,s::Space) -> Vector

returns the coefficients of f in the space s, which may not be the same as space(f).

coefficients(cfs::AbstractVector,fromspace::Space,tospace::Space) -> Vector

converts coefficients in fromspace to coefficients in tospace

ApproxFun.extrapolate — Function.

extrapolate(f::Fun,x)

returns an extrapolation of f from its domain to x.

ApproxFun.ncoefficients — Function.

ncoefficients(f::Fun) -> Integer

returns the number of coefficients of a fun

ApproxFun.points — Function.

points(f::Fun)

returns a grid of points that f can be transformed into values and back

points(s::Space,n::Integer)

returns a grid of approximately n points, for which a transform exists from values at the grid to coefficients in the space s.

ApproxFun.space — Function.

space(f::Fun)

returns the space of f

Base.values — Function.

values(f::Fun)

returns f evaluated at points(f)

Base.stride — Method.

stride(f::Fun)

returns the stride of the coefficients, checked numerically

Modify a Fun

ApproxFun.reverseorientation — Function.

reverseorientation(f::Fun)

return f on a reversed orientated contour.

ApproxFun.setdomain — Function.

setdomain(f::Fun,d::Domain)

returns f projected onto domain

Base.chop — Function.

chop(f::Fun,tol) -> Fun

reduces the number of coefficients by dropping the tail that is below the specified tolerance.

Operators

ApproxFun.Operator — Type.

Operator{T}

is an abstract type to represent linear operators between spaces.

BandedMatrices.bandwidths — Method.

bandwidths(op::Operator)

returns the bandwidth of op in the form (l,u), where l ≥ 0 represents the number of subdiagonals and u ≥ 0 represents the number of superdiagonals.

ApproxFun.domainspace — Function.

domainspace(op::Operator)

gives the domain space of op. That is, op*f will first convert f to a Fun in the space domainspace(op) before applying the operator.

ApproxFun.rangespace — Function.

rangespace(op::Operator)

gives the range space of op. That is, op*f will return a Fun in the space rangespace(op), provided f can be converted to a Fun in domainspace(op).

Base.getindex — Method.

op[f::Fun]

constructs the operator op*Multiplication(f), that is, it multiplies on the right by f first. Note that op*f is different: it applies op to f.

LinearAlgebra.qr — Method.

qr(A::Operator)

returns a cached QR factorization of the Operator A. The result QR enables solving of linear equations: if u=QR, then u approximately satisfies A*u = b.

LazyArrays.cache — Method.

cache(op::Operator)

Caches the entries of an operator, to speed up multiplying a Fun by the operator.

Inbuilt operators

ApproxFun.Conversion — Type.

Conversion(fromspace::Space,tospace::Space)

represents a conversion operator between fromspace and tospace, when available.

ApproxFun.Derivative — Type.

Derivative(sp::Space,k::Int) represents the k-th derivative on sp.

Derivative(sp::Space,k::Vector{Int}) represents a partial derivative on a multivariate space. For example,

Dx = Derivative(Chebyshev()^2,[1,0]) # ∂/∂x
Dy = Derivative(Chebyshev()^2,[0,1]) # ∂/∂y

Derivative(sp::Space) represents the first derivative Derivative(sp,1).

Derivative(k) represents the k-th derivative, acting on an unset space. Spaces will be inferred when applying or manipulating the operator.

Derivative() represents the first derivative on an unset space. Spaces will be inferred when applying or manipulating the operator.

ApproxFun.Dirichlet — Type.

Dirichlet(sp,k) is the operator associated with restricting the k-th derivative on the boundary for the space sp.

Dirichlet(sp) is the operator associated with restricting the the boundary for the space sp.

Dirichlet() is the operator associated with restricting on the the boundary.

ApproxFun.Evaluation — Type.

Evaluation(sp,x,k) is the functional associated with evaluating the k-th derivative at a point x for the space sp.

Evaluation(sp,x) is the functional associated with evaluating at a point x for the space sp.

Evaluation(x) is the functional associated with evaluating at a point x.

ApproxFun.Integral — Type.

Integral(sp::Space,k::Int) represents a k-th integral on sp. There is no guarantee on normalization.

Integral(sp::Space) represents the first integral Integral(sp,1).

Integral(k)represents thek`-th integral, acting on an unset space. Spaces will be inferred when applying or manipulating the operator.

Intergral() represents the first integral on an unset space. Spaces will be inferred when applying or manipulating the operator.

ApproxFun.Laplacian — Type.

Laplacian(sp::Space) represents the laplacian on space sp.

Laplacian() represents the laplacian on an unset space. Spaces will be inferred when applying or manipulating the operator.

ApproxFun.Multiplication — Type.

Multiplication(f::Fun,sp::Space) is the operator representing multiplication by f on functions in the space sp.

Multiplication(f::Fun) is the operator representing multiplication by f on an unset space of functions. Spaces will be inferred when applying or manipulating the operator.

ApproxFun.Neumann — Function.

Neumann(sp) is the operator associated with restricting the normal derivative on the boundary for the space sp. At the moment it is implemented as Dirichlet(sp,1).

`Neumann( is the operator associated with restricting the normal derivative on the boundary.

Bivariate

ApproxFun.LowRankFun — Type.

LowRankFun gives an approximation to a bivariate function in low rank form.